% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_67.R
\name{abany}
\alias{abany}
\title{abany}
\source{
General Social Survey https://gss.norc.org
}
\description{
Abortion if woman wants for any reason
}
\details{
Question 206. Please tell me whether or not you think it should be possible for a pregnant woman to obtaina legal abortion if: g. The woman wants it for any reason?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 44 \tab 921 \tab 7 \tab 558 \tab - \tab 1530 \cr
   1978 \tab - \tab 43 \tab 990 \tab 5 \tab 494 \tab - \tab 1532 \cr
   1980 \tab - \tab 59 \tab 828 \tab 3 \tab 578 \tab - \tab 1468 \cr
   1982 \tab - \tab 89 \tab 1082 \tab 11 \tab 678 \tab - \tab 1860 \cr
   1983 \tab - \tab 50 \tab 996 \tab 34 \tab 519 \tab - \tab 1599 \cr
   1984 \tab - \tab 42 \tab 872 \tab 11 \tab 548 \tab - \tab 1473 \cr
   1985 \tab - \tab 48 \tab 934 \tab 5 \tab 547 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 75 \tab 1051 \tab 14 \tab 679 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 37 \tab 598 \tab 4 \tab 338 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 39 \tab 590 \tab 5 \tab 399 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 36 \tab 496 \tab 4 \tab 381 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 35 \tab 546 \tab 7 \tab 405 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 59 \tab 552 \tab 6 \tab 458 \tab - \tab 1606 \cr
   1994 \tab 996 \tab 56 \tab 1039 \tab 6 \tab 895 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 94 \tab 1002 \tab 8 \tab 819 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 98 \tab 1050 \tab 6 \tab 728 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 87 \tab 1063 \tab 6 \tab 705 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 23 \tab 513 \tab 1 \tab 387 \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 35 \tab 507 \tab 10 \tab 346 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 50 \tab 1155 \tab 14 \tab 784 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 43 \tab 748 \tab 11 \tab 550 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 34 \tab 693 \tab 17 \tab 537 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 40 \tab 694 \tab 20 \tab 554 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 43 \tab 907 \tab 18 \tab 746 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 52 \tab 973 \tab 28 \tab 837 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 30 \tab 760 \tab 20 \tab 764 \tab - \tab 2348 \cr
   2021 \tab 2675 \tab 4 \tab 579 \tab 2 \tab 749 \tab 23 \tab 4032 \cr
   2022 \tab 2367 \tab 24 \tab 467 \tab 10 \tab 669 \tab 7 \tab 3544 \cr
   Total \tab 31440 \tab 1369 \tab 22606 \tab 293 \tab 16652 \tab 30 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
