% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_67.R
\name{abhlth}
\alias{abhlth}
\title{Woman's health seriously endangered}
\source{
General Social Survey https://gss.norc.org
}
\description{
abhlth
}
\details{
Question 206. Please tell me whether or not you think it should be possible for a pregnant woman to obtaina legal abortion if: c. The woman's own health is seriously endangered by the pregnancy?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab don't know \tab no \tab no answer \tab yes \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 66 \tab 201 \tab 8 \tab 1338 \tab - \tab - \tab 1613 \cr
   1973 \tab 26 \tab 114 \tab 2 \tab 1362 \tab - \tab - \tab 1504 \cr
   1974 \tab 32 \tab 110 \tab - \tab 1342 \tab - \tab - \tab 1484 \cr
   1975 \tab 38 \tab 135 \tab 3 \tab 1314 \tab - \tab - \tab 1490 \cr
   1976 \tab 28 \tab 135 \tab 7 \tab 1329 \tab - \tab - \tab 1499 \cr
   1977 \tab 34 \tab 141 \tab 8 \tab 1347 \tab - \tab - \tab 1530 \cr
   1978 \tab 36 \tab 140 \tab 4 \tab 1352 \tab - \tab - \tab 1532 \cr
   1980 \tab 37 \tab 141 \tab 2 \tab 1288 \tab - \tab - \tab 1468 \cr
   1982 \tab 55 \tab 173 \tab 5 \tab 1627 \tab - \tab - \tab 1860 \cr
   1983 \tab 47 \tab 156 \tab 32 \tab 1364 \tab - \tab - \tab 1599 \cr
   1984 \tab 31 \tab 150 \tab 11 \tab 1281 \tab - \tab - \tab 1473 \cr
   1985 \tab 40 \tab 159 \tab 3 \tab 1332 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab - \tab - \tab 1470 \tab - \tab 1470 \cr
   1987 \tab 66 \tab 211 \tab 7 \tab 1535 \tab - \tab - \tab 1819 \cr
   1988 \tab 32 \tab 106 \tab 3 \tab 836 \tab 504 \tab - \tab 1481 \cr
   1989 \tab 24 \tab 99 \tab 6 \tab 904 \tab 504 \tab - \tab 1537 \cr
   1990 \tab 27 \tab 73 \tab 3 \tab 814 \tab 455 \tab - \tab 1372 \cr
   1991 \tab 31 \tab 81 \tab 7 \tab 874 \tab 524 \tab - \tab 1517 \cr
   1993 \tab 43 \tab 105 \tab 4 \tab 923 \tab 531 \tab - \tab 1606 \cr
   1994 \tab 59 \tab 181 \tab 3 \tab 1753 \tab 996 \tab - \tab 2992 \cr
   1996 \tab 65 \tab 156 \tab 8 \tab 1694 \tab 981 \tab - \tab 2904 \cr
   1998 \tab 83 \tab 218 \tab 3 \tab 1578 \tab 950 \tab - \tab 2832 \cr
   2000 \tab 72 \tab 205 \tab 4 \tab 1580 \tab 956 \tab - \tab 2817 \cr
   2002 \tab 21 \tab 76 \tab 2 \tab 825 \tab 1841 \tab - \tab 2765 \cr
   2004 \tab 27 \tab 121 \tab 8 \tab 742 \tab 1914 \tab - \tab 2812 \cr
   2006 \tab 62 \tab 233 \tab 16 \tab 1692 \tab 2507 \tab - \tab 4510 \cr
   2008 \tab 57 \tab 139 \tab 10 \tab 1146 \tab 671 \tab - \tab 2023 \cr
   2010 \tab 35 \tab 158 \tab 22 \tab 1066 \tab 763 \tab - \tab 2044 \cr
   2012 \tab 58 \tab 146 \tab 21 \tab 1083 \tab 666 \tab - \tab 1974 \cr
   2014 \tab 46 \tab 195 \tab 21 \tab 1452 \tab 824 \tab - \tab 2538 \cr
   2016 \tab 53 \tab 202 \tab 25 \tab 1610 \tab 977 \tab - \tab 2867 \cr
   2018 \tab 43 \tab 151 \tab 14 \tab 1366 \tab 774 \tab - \tab 2348 \cr
   2021 \tab 9 \tab 141 \tab 1 \tab 1290 \tab 2574 \tab 17 \tab 4032 \cr
   2022 \tab 21 \tab 80 \tab 10 \tab 1062 \tab 2367 \tab 4 \tab 3544 \cr
   Total \tab 1404 \tab 4832 \tab 283 \tab 42101 \tab 23749 \tab 21 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
