% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_67.R
\name{abrape}
\alias{abrape}
\title{Pregnant as result of rape}
\source{
General Social Survey https://gss.norc.org
}
\description{
abrape
}
\details{
Question 206. Please tell me whether or not you think it should be possible for a pregnant woman to obtaina legal abortion if: e. She became pregnant as a result of rape?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab don't know \tab iap \tab no \tab no answer \tab yes \tab skipped on web \tab Total \cr
   1972 \tab 91 \tab 1 \tab 316 \tab 9 \tab 1196 \tab - \tab 1613 \cr
   1973 \tab 50 \tab - \tab 239 \tab 3 \tab 1212 \tab - \tab 1504 \cr
   1974 \tab 62 \tab - \tab 192 \tab 2 \tab 1228 \tab - \tab 1484 \cr
   1975 \tab 65 \tab - \tab 232 \tab 3 \tab 1190 \tab - \tab 1490 \cr
   1976 \tab 52 \tab - \tab 235 \tab 7 \tab 1205 \tab - \tab 1499 \cr
   1977 \tab 52 \tab - \tab 238 \tab 9 \tab 1231 \tab - \tab 1530 \cr
   1978 \tab 45 \tab - \tab 249 \tab 6 \tab 1232 \tab - \tab 1532 \cr
   1980 \tab 54 \tab - \tab 234 \tab 3 \tab 1177 \tab - \tab 1468 \cr
   1982 \tab 81 \tab - \tab 285 \tab 9 \tab 1485 \tab - \tab 1860 \cr
   1983 \tab 63 \tab - \tab 258 \tab 32 \tab 1246 \tab - \tab 1599 \cr
   1984 \tab 57 \tab - \tab 276 \tab 10 \tab 1130 \tab - \tab 1473 \cr
   1985 \tab 57 \tab - \tab 277 \tab 3 \tab 1197 \tab - \tab 1534 \cr
   1986 \tab - \tab 1470 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 79 \tab - \tab 353 \tab 11 \tab 1376 \tab - \tab 1819 \cr
   1988 \tab 50 \tab 504 \tab 175 \tab 3 \tab 749 \tab - \tab 1481 \cr
   1989 \tab 37 \tab 504 \tab 166 \tab 5 \tab 825 \tab - \tab 1537 \cr
   1990 \tab 39 \tab 455 \tab 133 \tab 5 \tab 740 \tab - \tab 1372 \cr
   1991 \tab 41 \tab 524 \tab 128 \tab 7 \tab 817 \tab - \tab 1517 \cr
   1993 \tab 46 \tab 531 \tab 175 \tab 4 \tab 850 \tab - \tab 1606 \cr
   1994 \tab 60 \tab 996 \tab 318 \tab 2 \tab 1616 \tab - \tab 2992 \cr
   1996 \tab 77 \tab 981 \tab 289 \tab 7 \tab 1550 \tab - \tab 2904 \cr
   1998 \tab 83 \tab 950 \tab 357 \tab 3 \tab 1439 \tab - \tab 2832 \cr
   2000 \tab 106 \tab 956 \tab 340 \tab 6 \tab 1409 \tab - \tab 2817 \cr
   2002 \tab 19 \tab 1841 \tab 184 \tab - \tab 721 \tab - \tab 2765 \cr
   2004 \tab 30 \tab 1914 \tab 203 \tab 14 \tab 651 \tab - \tab 2812 \cr
   2006 \tab 76 \tab 2507 \tab 429 \tab 15 \tab 1483 \tab - \tab 4510 \cr
   2008 \tab 62 \tab 671 \tab 298 \tab 10 \tab 982 \tab - \tab 2023 \cr
   2010 \tab 28 \tab 763 \tab 248 \tab 18 \tab 987 \tab - \tab 2044 \cr
   2012 \tab 59 \tab 666 \tab 289 \tab 19 \tab 941 \tab - \tab 1974 \cr
   2014 \tab 65 \tab 824 \tab 353 \tab 20 \tab 1276 \tab - \tab 2538 \cr
   2016 \tab 48 \tab 977 \tab 399 \tab 32 \tab 1411 \tab - \tab 2867 \cr
   2018 \tab 51 \tab 774 \tab 308 \tab 21 \tab 1194 \tab - \tab 2348 \cr
   2021 \tab 11 \tab 2574 \tab 219 \tab - \tab 1211 \tab 17 \tab 4032 \cr
   2022 \tab 19 \tab 2367 \tab 157 \tab 13 \tab 982 \tab 6 \tab 3544 \cr
   Total \tab 1815 \tab 23750 \tab 8552 \tab 311 \tab 37939 \tab 23 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
