% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_48.R
\name{affrmact}
\alias{affrmact}
\title{Favor preference in hiring blacks}
\source{
General Social Survey https://gss.norc.org
}
\description{
affrmact
}
\details{
Question 153. Some people say that because of past discrimination, blacks should be given preference in hiring and promotion. Others say that such preference in hiring and promotion of blacks is wrong because it discriminates against whites. What about your opinion - are you for or against preferential hiring and promotion of blacks? If favors: Do you favor preference in hiring and promotion strongly or not strongly? If opposes: Do you oppose preference in hiring and promotion strongly or not strongly?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab don't know \tab no answer \tab not strongly favors \tab not strongly opposes \tab strongly favors \tab strongly opposes \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1522 \tab 79 \tab 13 \tab 100 \tab 354 \tab 135 \tab 789 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 117 \tab 22 \tab 127 \tab 485 \tab 183 \tab 1026 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 145 \tab 18 \tab 117 \tab 425 \tab 142 \tab 1024 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 119 \tab 14 \tab 130 \tab 480 \tab 201 \tab 933 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 22 \tab 3 \tab 58 \tab 249 \tab 84 \tab 492 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 38 \tab 4 \tab 60 \tab 244 \tab 86 \tab 474 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 75 \tab 13 \tab 136 \tab 525 \tab 188 \tab 1055 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 65 \tab 9 \tab 87 \tab 349 \tab 133 \tab 686 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 66 \tab 14 \tab 114 \tab 377 \tab 160 \tab 699 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 63 \tab 18 \tab 99 \tab 354 \tab 139 \tab 629 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 67 \tab 16 \tab 130 \tab 475 \tab 173 \tab 814 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 87 \tab 24 \tab 151 \tab 545 \tab 268 \tab 813 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 88 \tab 24 \tab 132 \tab 443 \tab 230 \tab 642 \tab - \tab 2348 \cr
   2021 \tab 1378 \tab 22 \tab 4 \tab 299 \tab 818 \tab 349 \tab 1162 \tab - \tab 4032 \cr
   2022 \tab 1168 \tab 119 \tab 15 \tab 248 \tab 667 \tab 395 \tab 919 \tab 13 \tab 3544 \cr
   Total \tab 47193 \tab 1172 \tab 211 \tab 1988 \tab 6790 \tab 2866 \tab 12157 \tab 13 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} strongly favors
\item \code{2} not strongly favors
\item \code{3} not strongly opposes
\item \code{4} strongly opposes
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
