% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_09.R
\name{age}
\alias{age}
\title{age}
\source{
General Social Survey https://gss.norc.org
}
\description{
Age of respondent
}
\details{
Question 13. Respondent's age
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 76 \tab 77 \tab 78 \tab 79 \tab 80 \tab 81 \tab 82 \tab 83 \tab 84 \tab 85 \tab 87 \tab 88 \tab 89 or older \tab no answer \tab 86 \tab don't know \tab iap \tab Total \cr
   1972 \tab 5 \tab 24 \tab 34 \tab 37 \tab 38 \tab 31 \tab 31 \tab 40 \tab 48 \tab 37 \tab 30 \tab 45 \tab 50 \tab 25 \tab 25 \tab 26 \tab 32 \tab 29 \tab 28 \tab 18 \tab 23 \tab 21 \tab 16 \tab 22 \tab 29 \tab 31 \tab 23 \tab 42 \tab 31 \tab 30 \tab 38 \tab 36 \tab 28 \tab 26 \tab 32 \tab 29 \tab 24 \tab 26 \tab 31 \tab 26 \tab 24 \tab 25 \tab 34 \tab 21 \tab 15 \tab 23 \tab 27 \tab 24 \tab 11 \tab 13 \tab 20 \tab 16 \tab 17 \tab 9 \tab 19 \tab 12 \tab 17 \tab 13 \tab 12 \tab 10 \tab 12 \tab 4 \tab 6 \tab 5 \tab 6 \tab 4 \tab 3 \tab 3 \tab 2 \tab 2 \tab 2 \tab 5 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 6 \tab 29 \tab 29 \tab 32 \tab 42 \tab 33 \tab 43 \tab 32 \tab 33 \tab 40 \tab 30 \tab 34 \tab 29 \tab 29 \tab 22 \tab 21 \tab 35 \tab 31 \tab 41 \tab 28 \tab 32 \tab 27 \tab 24 \tab 23 \tab 24 \tab 24 \tab 21 \tab 22 \tab 32 \tab 23 \tab 24 \tab 30 \tab 20 \tab 28 \tab 29 \tab 20 \tab 30 \tab 18 \tab 33 \tab 19 \tab 29 \tab 28 \tab 21 \tab 16 \tab 23 \tab 20 \tab 24 \tab 17 \tab 24 \tab 16 \tab 17 \tab 15 \tab 17 \tab 14 \tab 12 \tab 11 \tab 25 \tab 7 \tab 7 \tab 10 \tab 6 \tab 8 \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 1504 \cr
   1974 \tab 10 \tab 21 \tab 25 \tab 33 \tab 29 \tab 50 \tab 34 \tab 28 \tab 41 \tab 43 \tab 36 \tab 30 \tab 35 \tab 38 \tab 30 \tab 22 \tab 24 \tab 26 \tab 19 \tab 26 \tab 35 \tab 31 \tab 26 \tab 15 \tab 32 \tab 24 \tab 28 \tab 22 \tab 25 \tab 20 \tab 28 \tab 15 \tab 23 \tab 32 \tab 27 \tab 21 \tab 26 \tab 19 \tab 24 \tab 19 \tab 26 \tab 25 \tab 23 \tab 22 \tab 10 \tab 17 \tab 11 \tab 25 \tab 18 \tab 19 \tab 17 \tab 27 \tab 10 \tab 17 \tab 21 \tab 12 \tab 13 \tab 8 \tab 11 \tab 5 \tab 8 \tab 6 \tab 8 \tab 6 \tab 1 \tab 6 \tab 2 \tab 2 \tab 1 \tab 1 \tab 5 \tab 6 \tab 3 \tab - \tab - \tab 1484 \cr
   1975 \tab 10 \tab 38 \tab 33 \tab 26 \tab 31 \tab 35 \tab 42 \tab 46 \tab 36 \tab 38 \tab 36 \tab 34 \tab 30 \tab 19 \tab 37 \tab 35 \tab 26 \tab 24 \tab 25 \tab 17 \tab 30 \tab 21 \tab 33 \tab 28 \tab 21 \tab 33 \tab 18 \tab 26 \tab 31 \tab 14 \tab 23 \tab 21 \tab 31 \tab 19 \tab 26 \tab 21 \tab 24 \tab 21 \tab 16 \tab 20 \tab 24 \tab 21 \tab 19 \tab 13 \tab 26 \tab 18 \tab 18 \tab 19 \tab 25 \tab 21 \tab 16 \tab 21 \tab 28 \tab 8 \tab 11 \tab 10 \tab 12 \tab 3 \tab 12 \tab 13 \tab 6 \tab 7 \tab 7 \tab 9 \tab 4 \tab 4 \tab 3 \tab 4 \tab 1 \tab 1 \tab 3 \tab 5 \tab 3 \tab - \tab - \tab 1490 \cr
   1976 \tab 8 \tab 34 \tab 32 \tab 28 \tab 26 \tab 34 \tab 41 \tab 35 \tab 30 \tab 39 \tab 42 \tab 39 \tab 35 \tab 24 \tab 40 \tab 27 \tab 37 \tab 23 \tab 34 \tab 25 \tab 26 \tab 28 \tab 18 \tab 27 \tab 20 \tab 16 \tab 15 \tab 22 \tab 22 \tab 19 \tab 20 \tab 20 \tab 13 \tab 21 \tab 27 \tab 32 \tab 18 \tab 23 \tab 22 \tab 18 \tab 29 \tab 19 \tab 22 \tab 27 \tab 29 \tab 16 \tab 24 \tab 16 \tab 17 \tab 26 \tab 11 \tab 13 \tab 20 \tab 11 \tab 16 \tab 16 \tab 9 \tab 10 \tab 17 \tab 10 \tab 13 \tab 10 \tab 9 \tab 5 \tab 4 \tab 4 \tab 7 \tab 4 \tab 5 \tab 1 \tab 7 \tab 6 \tab 6 \tab - \tab - \tab 1499 \cr
   1977 \tab 3 \tab 26 \tab 38 \tab 25 \tab 36 \tab 36 \tab 28 \tab 34 \tab 48 \tab 35 \tab 34 \tab 25 \tab 40 \tab 25 \tab 20 \tab 39 \tab 35 \tab 29 \tab 33 \tab 32 \tab 25 \tab 25 \tab 26 \tab 27 \tab 37 \tab 20 \tab 21 \tab 30 \tab 21 \tab 18 \tab 24 \tab 31 \tab 28 \tab 21 \tab 19 \tab 31 \tab 28 \tab 25 \tab 24 \tab 30 \tab 33 \tab 28 \tab 19 \tab 18 \tab 19 \tab 27 \tab 15 \tab 19 \tab 16 \tab 19 \tab 19 \tab 14 \tab 11 \tab 18 \tab 13 \tab 18 \tab 17 \tab 10 \tab 10 \tab 11 \tab 5 \tab 6 \tab 3 \tab 6 \tab 3 \tab 3 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab - \tab 1530 \cr
   1978 \tab 7 \tab 25 \tab 28 \tab 32 \tab 29 \tab 42 \tab 41 \tab 40 \tab 51 \tab 35 \tab 40 \tab 37 \tab 39 \tab 44 \tab 44 \tab 34 \tab 34 \tab 35 \tab 22 \tab 27 \tab 27 \tab 28 \tab 33 \tab 23 \tab 27 \tab 28 \tab 21 \tab 13 \tab 13 \tab 17 \tab 20 \tab 22 \tab 21 \tab 25 \tab 23 \tab 22 \tab 22 \tab 17 \tab 20 \tab 32 \tab 26 \tab 20 \tab 20 \tab 16 \tab 18 \tab 13 \tab 18 \tab 22 \tab 18 \tab 13 \tab 16 \tab 11 \tab 14 \tab 20 \tab 15 \tab 15 \tab 12 \tab 11 \tab 11 \tab 12 \tab 8 \tab 14 \tab 7 \tab 9 \tab 3 \tab 4 \tab 4 \tab 1 \tab 1 \tab 2 \tab 5 \tab 7 \tab 6 \tab - \tab - \tab 1532 \cr
   1980 \tab 2 \tab 28 \tab 26 \tab 31 \tab 38 \tab 30 \tab 39 \tab 22 \tab 27 \tab 42 \tab 34 \tab 38 \tab 38 \tab 34 \tab 38 \tab 46 \tab 31 \tab 28 \tab 32 \tab 35 \tab 19 \tab 18 \tab 25 \tab 21 \tab 20 \tab 15 \tab 26 \tab 20 \tab 17 \tab 19 \tab 19 \tab 28 \tab 20 \tab 20 \tab 16 \tab 20 \tab 21 \tab 20 \tab 22 \tab 27 \tab 24 \tab 22 \tab 18 \tab 19 \tab 16 \tab 24 \tab 16 \tab 26 \tab 23 \tab 26 \tab 20 \tab 18 \tab 18 \tab 20 \tab 11 \tab 9 \tab 9 \tab 9 \tab 12 \tab 13 \tab 4 \tab 5 \tab 5 \tab 6 \tab 6 \tab 6 \tab 5 \tab 2 \tab 4 \tab 2 \tab 5 \tab 9 \tab 4 \tab - \tab - \tab 1468 \cr
   1982 \tab 6 \tab 33 \tab 25 \tab 40 \tab 35 \tab 45 \tab 41 \tab 57 \tab 47 \tab 61 \tab 51 \tab 55 \tab 51 \tab 47 \tab 38 \tab 31 \tab 37 \tab 38 \tab 32 \tab 39 \tab 33 \tab 27 \tab 30 \tab 29 \tab 25 \tab 24 \tab 28 \tab 17 \tab 26 \tab 24 \tab 24 \tab 27 \tab 24 \tab 20 \tab 30 \tab 22 \tab 31 \tab 21 \tab 38 \tab 25 \tab 31 \tab 29 \tab 24 \tab 31 \tab 19 \tab 25 \tab 22 \tab 20 \tab 19 \tab 23 \tab 29 \tab 21 \tab 21 \tab 22 \tab 13 \tab 21 \tab 15 \tab 15 \tab 12 \tab 14 \tab 9 \tab 12 \tab 10 \tab 8 \tab 8 \tab 5 \tab 6 \tab 5 \tab 4 \tab 3 \tab 10 \tab 15 \tab 5 \tab - \tab - \tab 1860 \cr
   1983 \tab 10 \tab 23 \tab 12 \tab 21 \tab 28 \tab 34 \tab 36 \tab 45 \tab 56 \tab 59 \tab 48 \tab 43 \tab 33 \tab 50 \tab 41 \tab 36 \tab 38 \tab 38 \tab 31 \tab 37 \tab 32 \tab 25 \tab 25 \tab 23 \tab 26 \tab 22 \tab 29 \tab 25 \tab 23 \tab 20 \tab 16 \tab 21 \tab 17 \tab 14 \tab 17 \tab 31 \tab 24 \tab 11 \tab 30 \tab 21 \tab 32 \tab 16 \tab 29 \tab 21 \tab 19 \tab 19 \tab 18 \tab 26 \tab 13 \tab 27 \tab 15 \tab 16 \tab 17 \tab 14 \tab 20 \tab 14 \tab 12 \tab 10 \tab 10 \tab 10 \tab 9 \tab 8 \tab 5 \tab 11 \tab 7 \tab 6 \tab 5 \tab 2 \tab 1 \tab 1 \tab 6 \tab 7 \tab 2 \tab - \tab - \tab 1599 \cr
   1984 \tab 2 \tab 20 \tab 23 \tab 37 \tab 29 \tab 50 \tab 47 \tab 39 \tab 32 \tab 44 \tab 37 \tab 33 \tab 28 \tab 28 \tab 37 \tab 39 \tab 31 \tab 37 \tab 41 \tab 38 \tab 31 \tab 24 \tab 28 \tab 25 \tab 21 \tab 25 \tab 23 \tab 20 \tab 18 \tab 19 \tab 22 \tab 21 \tab 8 \tab 16 \tab 21 \tab 14 \tab 17 \tab 19 \tab 19 \tab 19 \tab 24 \tab 16 \tab 27 \tab 20 \tab 20 \tab 15 \tab 13 \tab 18 \tab 14 \tab 25 \tab 24 \tab 9 \tab 18 \tab 12 \tab 14 \tab 6 \tab 15 \tab 9 \tab 14 \tab 13 \tab 6 \tab 4 \tab 6 \tab 12 \tab 8 \tab 5 \tab 4 \tab 2 \tab 1 \tab 2 \tab 7 \tab 6 \tab 2 \tab - \tab - \tab 1473 \cr
   1985 \tab 4 \tab 15 \tab 30 \tab 25 \tab 26 \tab 33 \tab 24 \tab 41 \tab 37 \tab 47 \tab 40 \tab 28 \tab 37 \tab 30 \tab 49 \tab 33 \tab 31 \tab 32 \tab 31 \tab 36 \tab 32 \tab 30 \tab 30 \tab 25 \tab 23 \tab 22 \tab 20 \tab 18 \tab 25 \tab 16 \tab 18 \tab 23 \tab 21 \tab 34 \tab 15 \tab 20 \tab 20 \tab 13 \tab 25 \tab 20 \tab 29 \tab 27 \tab 20 \tab 20 \tab 23 \tab 27 \tab 17 \tab 24 \tab 16 \tab 23 \tab 15 \tab 21 \tab 22 \tab 16 \tab 12 \tab 12 \tab 11 \tab 13 \tab 10 \tab 11 \tab 12 \tab 8 \tab 8 \tab 12 \tab 6 \tab 7 \tab 6 \tab 7 \tab 7 \tab - \tab 5 \tab 7 \tab 1 \tab - \tab - \tab 1534 \cr
   1986 \tab 1 \tab 18 \tab 19 \tab 22 \tab 26 \tab 22 \tab 28 \tab 43 \tab 31 \tab 39 \tab 42 \tab 35 \tab 42 \tab 37 \tab 45 \tab 26 \tab 39 \tab 32 \tab 32 \tab 31 \tab 32 \tab 37 \tab 27 \tab 29 \tab 28 \tab 28 \tab 25 \tab 21 \tab 25 \tab 14 \tab 16 \tab 19 \tab 22 \tab 25 \tab 12 \tab 14 \tab 20 \tab 15 \tab 19 \tab 15 \tab 10 \tab 20 \tab 21 \tab 14 \tab 27 \tab 18 \tab 15 \tab 20 \tab 25 \tab 22 \tab 20 \tab 11 \tab 22 \tab 14 \tab 15 \tab 12 \tab 16 \tab 6 \tab 11 \tab 13 \tab 15 \tab 12 \tab 8 \tab 10 \tab 7 \tab 4 \tab 7 \tab 3 \tab 3 \tab - \tab 5 \tab 7 \tab 4 \tab - \tab - \tab 1470 \cr
   1987 \tab 5 \tab 30 \tab 24 \tab 31 \tab 32 \tab 47 \tab 34 \tab 38 \tab 37 \tab 43 \tab 44 \tab 44 \tab 53 \tab 42 \tab 41 \tab 35 \tab 62 \tab 46 \tab 43 \tab 37 \tab 40 \tab 41 \tab 41 \tab 28 \tab 27 \tab 32 \tab 27 \tab 43 \tab 38 \tab 19 \tab 30 \tab 19 \tab 21 \tab 26 \tab 20 \tab 30 \tab 18 \tab 26 \tab 13 \tab 20 \tab 23 \tab 20 \tab 19 \tab 15 \tab 26 \tab 22 \tab 23 \tab 28 \tab 26 \tab 18 \tab 22 \tab 20 \tab 20 \tab 23 \tab 17 \tab 22 \tab 12 \tab 15 \tab 12 \tab 14 \tab 16 \tab 8 \tab 7 \tab 7 \tab 4 \tab 7 \tab 5 \tab 6 \tab 3 \tab 4 \tab 8 \tab 12 \tab 8 \tab - \tab - \tab 1819 \cr
   1988 \tab 3 \tab 24 \tab 24 \tab 28 \tab 31 \tab 31 \tab 29 \tab 40 \tab 36 \tab 38 \tab 42 \tab 30 \tab 36 \tab 35 \tab 24 \tab 30 \tab 28 \tab 42 \tab 28 \tab 40 \tab 39 \tab 22 \tab 31 \tab 33 \tab 34 \tab 28 \tab 33 \tab 28 \tab 16 \tab 18 \tab 11 \tab 22 \tab 15 \tab 18 \tab 12 \tab 14 \tab 10 \tab 15 \tab 17 \tab 10 \tab 19 \tab 14 \tab 22 \tab 22 \tab 20 \tab 19 \tab 22 \tab 14 \tab 19 \tab 18 \tab 23 \tab 20 \tab 18 \tab 18 \tab 15 \tab 12 \tab 15 \tab 17 \tab 16 \tab 10 \tab 18 \tab 13 \tab 3 \tab 5 \tab 3 \tab 8 \tab 4 \tab 5 \tab 4 \tab 2 \tab 11 \tab 4 \tab 3 \tab - \tab - \tab 1481 \cr
   1989 \tab 5 \tab 26 \tab 21 \tab 33 \tab 23 \tab 29 \tab 28 \tab 35 \tab 23 \tab 33 \tab 47 \tab 36 \tab 34 \tab 42 \tab 36 \tab 39 \tab 29 \tab 32 \tab 43 \tab 29 \tab 40 \tab 28 \tab 29 \tab 29 \tab 35 \tab 25 \tab 23 \tab 28 \tab 22 \tab 28 \tab 24 \tab 24 \tab 27 \tab 31 \tab 15 \tab 17 \tab 17 \tab 24 \tab 16 \tab 14 \tab 15 \tab 13 \tab 21 \tab 18 \tab 16 \tab 20 \tab 12 \tab 24 \tab 24 \tab 19 \tab 26 \tab 18 \tab 18 \tab 8 \tab 22 \tab 8 \tab 15 \tab 17 \tab 11 \tab 14 \tab 12 \tab 8 \tab 6 \tab 9 \tab 10 \tab 4 \tab 2 \tab 5 \tab 3 \tab 5 \tab 9 \tab 4 \tab 2 \tab - \tab - \tab 1537 \cr
   1990 \tab 4 \tab 15 \tab 22 \tab 19 \tab 37 \tab 19 \tab 26 \tab 27 \tab 16 \tab 24 \tab 33 \tab 34 \tab 35 \tab 32 \tab 36 \tab 35 \tab 36 \tab 33 \tab 35 \tab 35 \tab 40 \tab 21 \tab 33 \tab 31 \tab 26 \tab 34 \tab 25 \tab 23 \tab 20 \tab 25 \tab 19 \tab 18 \tab 14 \tab 14 \tab 15 \tab 14 \tab 15 \tab 8 \tab 21 \tab 15 \tab 13 \tab 12 \tab 15 \tab 17 \tab 18 \tab 17 \tab 17 \tab 18 \tab 18 \tab 16 \tab 17 \tab 19 \tab 14 \tab 8 \tab 16 \tab 20 \tab 21 \tab 9 \tab 9 \tab 14 \tab 12 \tab 11 \tab 6 \tab 9 \tab 5 \tab 9 \tab 6 \tab 4 \tab 2 \tab 4 \tab 10 \tab - \tab 2 \tab - \tab - \tab 1372 \cr
   1991 \tab 3 \tab 9 \tab 18 \tab 38 \tab 35 \tab 28 \tab 24 \tab 28 \tab 33 \tab 29 \tab 33 \tab 27 \tab 35 \tab 35 \tab 44 \tab 38 \tab 35 \tab 55 \tab 35 \tab 33 \tab 37 \tab 35 \tab 36 \tab 38 \tab 30 \tab 33 \tab 30 \tab 22 \tab 12 \tab 28 \tab 29 \tab 22 \tab 20 \tab 15 \tab 20 \tab 14 \tab 18 \tab 16 \tab 12 \tab 17 \tab 18 \tab 15 \tab 16 \tab 20 \tab 17 \tab 19 \tab 15 \tab 18 \tab 13 \tab 18 \tab 18 \tab 17 \tab 15 \tab 20 \tab 19 \tab 24 \tab 17 \tab 10 \tab 13 \tab 12 \tab 11 \tab 7 \tab 11 \tab 9 \tab 13 \tab 5 \tab 5 \tab 6 \tab 4 \tab 2 \tab 8 \tab 3 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 5 \tab 18 \tab 20 \tab 24 \tab 16 \tab 28 \tab 26 \tab 35 \tab 27 \tab 23 \tab 46 \tab 33 \tab 39 \tab 36 \tab 30 \tab 36 \tab 28 \tab 44 \tab 45 \tab 38 \tab 44 \tab 41 \tab 40 \tab 39 \tab 32 \tab 40 \tab 29 \tab 33 \tab 33 \tab 36 \tab 20 \tab 29 \tab 24 \tab 27 \tab 22 \tab 18 \tab 19 \tab 23 \tab 13 \tab 19 \tab 25 \tab 16 \tab 17 \tab 12 \tab 18 \tab 19 \tab 14 \tab 18 \tab 20 \tab 11 \tab 18 \tab 20 \tab 12 \tab 17 \tab 20 \tab 20 \tab 19 \tab 18 \tab 14 \tab 15 \tab 14 \tab 7 \tab 6 \tab 9 \tab 12 \tab 3 \tab 4 \tab 4 \tab 6 \tab 3 \tab 7 \tab 5 \tab 5 \tab - \tab - \tab 1606 \cr
   1994 \tab 9 \tab 27 \tab 34 \tab 29 \tab 31 \tab 42 \tab 51 \tab 62 \tab 55 \tab 57 \tab 68 \tab 56 \tab 69 \tab 74 \tab 83 \tab 78 \tab 88 \tab 76 \tab 80 \tab 60 \tab 80 \tab 65 \tab 59 \tab 71 \tab 59 \tab 69 \tab 65 \tab 48 \tab 60 \tab 71 \tab 55 \tab 66 \tab 55 \tab 34 \tab 56 \tab 45 \tab 44 \tab 32 \tab 38 \tab 40 \tab 29 \tab 43 \tab 32 \tab 26 \tab 25 \tab 42 \tab 23 \tab 27 \tab 28 \tab 30 \tab 32 \tab 30 \tab 28 \tab 32 \tab 25 \tab 26 \tab 31 \tab 28 \tab 32 \tab 20 \tab 21 \tab 20 \tab 21 \tab 11 \tab 16 \tab 9 \tab 9 \tab 12 \tab 6 \tab 4 \tab 20 \tab 6 \tab 7 \tab - \tab - \tab 2992 \cr
   1996 \tab 7 \tab 34 \tab 40 \tab 37 \tab 45 \tab 58 \tab 66 \tab 66 \tab 58 \tab 53 \tab 68 \tab 59 \tab 72 \tab 60 \tab 75 \tab 74 \tab 64 \tab 61 \tab 61 \tab 89 \tab 82 \tab 58 \tab 68 \tab 63 \tab 67 \tab 74 \tab 53 \tab 54 \tab 52 \tab 57 \tab 66 \tab 69 \tab 48 \tab 49 \tab 40 \tab 55 \tab 45 \tab 31 \tab 41 \tab 25 \tab 27 \tab 33 \tab 33 \tab 26 \tab 38 \tab 29 \tab 24 \tab 27 \tab 25 \tab 21 \tab 31 \tab 24 \tab 39 \tab 13 \tab 27 \tab 17 \tab 27 \tab 18 \tab 16 \tab 24 \tab 17 \tab 15 \tab 16 \tab 8 \tab 11 \tab 11 \tab 9 \tab 5 \tab 7 \tab 7 \tab 18 \tab 6 \tab 11 \tab - \tab - \tab 2904 \cr
   1998 \tab 6 \tab 36 \tab 36 \tab 48 \tab 37 \tab 42 \tab 37 \tab 50 \tab 61 \tab 50 \tab 77 \tab 57 \tab 53 \tab 56 \tab 66 \tab 74 \tab 83 \tab 79 \tab 71 \tab 71 \tab 63 \tab 71 \tab 70 \tab 63 \tab 63 \tab 61 \tab 67 \tab 54 \tab 52 \tab 49 \tab 51 \tab 50 \tab 42 \tab 72 \tab 40 \tab 34 \tab 37 \tab 45 \tab 42 \tab 33 \tab 34 \tab 26 \tab 31 \tab 26 \tab 33 \tab 22 \tab 28 \tab 27 \tab 23 \tab 29 \tab 24 \tab 24 \tab 30 \tab 24 \tab 26 \tab 41 \tab 20 \tab 27 \tab 20 \tab 21 \tab 12 \tab 18 \tab 15 \tab 12 \tab 13 \tab 14 \tab 7 \tab 13 \tab 12 \tab 6 \tab 12 \tab 4 \tab 9 \tab - \tab - \tab 2832 \cr
   2000 \tab 7 \tab 44 \tab 40 \tab 37 \tab 39 \tab 48 \tab 52 \tab 53 \tab 58 \tab 47 \tab 45 \tab 56 \tab 62 \tab 62 \tab 77 \tab 45 \tab 55 \tab 55 \tab 62 \tab 63 \tab 76 \tab 53 \tab 73 \tab 60 \tab 73 \tab 70 \tab 70 \tab 69 \tab 55 \tab 54 \tab 62 \tab 48 \tab 51 \tab 42 \tab 58 \tab 51 \tab 46 \tab 26 \tab 42 \tab 29 \tab 29 \tab 34 \tab 32 \tab 38 \tab 29 \tab 29 \tab 20 \tab 18 \tab 32 \tab 27 \tab 28 \tab 20 \tab 22 \tab 21 \tab 28 \tab 28 \tab 20 \tab 19 \tab 20 \tab 20 \tab 22 \tab 22 \tab 17 \tab 17 \tab 15 \tab 17 \tab 17 \tab 15 \tab 5 \tab 5 \tab 17 \tab 8 \tab 11 \tab - \tab - \tab 2817 \cr
   2002 \tab 4 \tab 25 \tab 32 \tab 40 \tab 46 \tab 68 \tab 58 \tab 61 \tab 54 \tab 48 \tab 51 \tab 43 \tab 61 \tab 53 \tab 62 \tab 76 \tab 57 \tab 59 \tab 55 \tab 64 \tab 49 \tab 54 \tab 60 \tab 54 \tab 44 \tab 63 \tab 47 \tab 51 \tab 70 \tab 52 \tab 61 \tab 43 \tab 47 \tab 52 \tab 58 \tab 45 \tab 45 \tab 39 \tab 47 \tab 35 \tab 33 \tab 33 \tab 30 \tab 31 \tab 32 \tab 34 \tab 38 \tab 30 \tab 20 \tab 30 \tab 26 \tab 27 \tab 34 \tab 36 \tab 23 \tab 22 \tab 24 \tab 26 \tab 31 \tab 19 \tab 15 \tab 9 \tab 16 \tab 13 \tab 11 \tab 15 \tab 14 \tab 11 \tab 10 \tab 6 \tab 13 \tab 14 \tab 6 \tab - \tab - \tab 2765 \cr
   2004 \tab 27 \tab 33 \tab 32 \tab 38 \tab 36 \tab 47 \tab 50 \tab 55 \tab 59 \tab 52 \tab 69 \tab 43 \tab 55 \tab 52 \tab 52 \tab 58 \tab 70 \tab 62 \tab 58 \tab 45 \tab 54 \tab 51 \tab 73 \tab 58 \tab 49 \tab 60 \tab 68 \tab 57 \tab 58 \tab 62 \tab 57 \tab 61 \tab 59 \tab 53 \tab 58 \tab 40 \tab 36 \tab 50 \tab 56 \tab 53 \tab 36 \tab 39 \tab 44 \tab 43 \tab 43 \tab 36 \tab 26 \tab 32 \tab 29 \tab 41 \tab 22 \tab 22 \tab 22 \tab 22 \tab 18 \tab 12 \tab 23 \tab 25 \tab 14 \tab 8 \tab 17 \tab 22 \tab 9 \tab 15 \tab 15 \tab 11 \tab 7 \tab 11 \tab 4 \tab 7 \tab 14 \tab 9 \tab 8 \tab - \tab - \tab 2812 \cr
   2006 \tab 19 \tab 47 \tab 50 \tab 52 \tab 72 \tab 63 \tab 61 \tab 74 \tab 83 \tab 86 \tab 80 \tab 79 \tab 84 \tab 67 \tab 97 \tab 88 \tab 91 \tab 80 \tab 105 \tab 89 \tab 94 \tab 95 \tab 95 \tab 94 \tab 99 \tab 95 \tab 100 \tab 79 \tab 90 \tab 110 \tab 109 \tab 97 \tab 93 \tab 76 \tab 78 \tab 89 \tab 96 \tab 86 \tab 68 \tab 58 \tab 81 \tab 78 \tab 75 \tab 57 \tab 58 \tab 62 \tab 56 \tab 42 \tab 53 \tab 45 \tab 41 \tab 42 \tab 41 \tab 39 \tab 48 \tab 29 \tab 33 \tab 39 \tab 31 \tab 33 \tab 28 \tab 22 \tab 27 \tab 25 \tab 21 \tab 19 \tab 17 \tab 18 \tab 7 \tab 11 \tab 29 \tab 18 \tab 17 \tab - \tab - \tab 4510 \cr
   2008 \tab 6 \tab 31 \tab 28 \tab 21 \tab 27 \tab 37 \tab 34 \tab 26 \tab 41 \tab 37 \tab 40 \tab 31 \tab 27 \tab 47 \tab 22 \tab 36 \tab 40 \tab 41 \tab 45 \tab 44 \tab 38 \tab 33 \tab 39 \tab 36 \tab 28 \tab 35 \tab 40 \tab 34 \tab 41 \tab 49 \tab 38 \tab 41 \tab 51 \tab 43 \tab 37 \tab 42 \tab 38 \tab 33 \tab 42 \tab 41 \tab 24 \tab 20 \tab 40 \tab 36 \tab 34 \tab 19 \tab 25 \tab 31 \tab 23 \tab 27 \tab 16 \tab 21 \tab 21 \tab 16 \tab 19 \tab 11 \tab 23 \tab 18 \tab 17 \tab 11 \tab 17 \tab 12 \tab 15 \tab 13 \tab 12 \tab 8 \tab 7 \tab 3 \tab 6 \tab 1 \tab 21 \tab 10 \tab 6 \tab - \tab - \tab 2023 \cr
   2010 \tab 10 \tab 24 \tab 24 \tab 35 \tab 19 \tab 38 \tab 31 \tab 48 \tab 29 \tab 42 \tab 30 \tab 45 \tab 40 \tab 42 \tab 33 \tab 36 \tab 32 \tab 45 \tab 30 \tab 39 \tab 31 \tab 35 \tab 36 \tab 38 \tab 34 \tab 42 \tab 39 \tab 38 \tab 40 \tab 29 \tab 27 \tab 45 \tab 43 \tab 39 \tab 37 \tab 37 \tab 40 \tab 32 \tab 40 \tab 31 \tab 38 \tab 41 \tab 44 \tab 27 \tab 39 \tab 35 \tab 19 \tab 28 \tab 20 \tab 37 \tab 27 \tab 26 \tab 24 \tab 15 \tab 18 \tab 8 \tab 15 \tab 15 \tab 17 \tab 14 \tab 8 \tab 11 \tab 17 \tab 10 \tab 13 \tab 12 \tab 12 \tab 6 \tab 4 \tab 9 \tab 20 \tab 3 \tab 7 \tab - \tab - \tab 2044 \cr
   2012 \tab 12 \tab 20 \tab 19 \tab 31 \tab 31 \tab 30 \tab 27 \tab 31 \tab 29 \tab 32 \tab 41 \tab 28 \tab 47 \tab 35 \tab 47 \tab 39 \tab 34 \tab 35 \tab 33 \tab 40 \tab 37 \tab 42 \tab 30 \tab 40 \tab 43 \tab 29 \tab 29 \tab 35 \tab 34 \tab 36 \tab 38 \tab 43 \tab 32 \tab 40 \tab 37 \tab 32 \tab 33 \tab 47 \tab 29 \tab 31 \tab 25 \tab 31 \tab 27 \tab 37 \tab 31 \tab 35 \tab 31 \tab 38 \tab 19 \tab 23 \tab 24 \tab 20 \tab 17 \tab 29 \tab 15 \tab 20 \tab 17 \tab 16 \tab 17 \tab 15 \tab 15 \tab 11 \tab 10 \tab 12 \tab 10 \tab 6 \tab 13 \tab 5 \tab 10 \tab 7 \tab 15 \tab 5 \tab 10 \tab - \tab - \tab 1974 \cr
   2014 \tab 6 \tab 25 \tab 26 \tab 24 \tab 28 \tab 30 \tab 31 \tab 48 \tab 47 \tab 41 \tab 31 \tab 51 \tab 57 \tab 49 \tab 55 \tab 47 \tab 46 \tab 40 \tab 40 \tab 54 \tab 47 \tab 52 \tab 46 \tab 54 \tab 35 \tab 54 \tab 39 \tab 41 \tab 34 \tab 43 \tab 32 \tab 39 \tab 54 \tab 45 \tab 37 \tab 60 \tab 53 \tab 52 \tab 60 \tab 43 \tab 60 \tab 47 \tab 46 \tab 38 \tab 44 \tab 42 \tab 38 \tab 40 \tab 35 \tab 41 \tab 21 \tab 23 \tab 32 \tab 28 \tab 20 \tab 22 \tab 25 \tab 21 \tab 24 \tab 17 \tab 28 \tab 26 \tab 16 \tab 14 \tab 8 \tab 11 \tab 13 \tab 6 \tab 8 \tab 11 \tab 19 \tab 9 \tab 9 \tab - \tab - \tab 2538 \cr
   2016 \tab 7 \tab 33 \tab 26 \tab 33 \tab 44 \tab 49 \tab 35 \tab 56 \tab 42 \tab 58 \tab 42 \tab 56 \tab 54 \tab 57 \tab 42 \tab 54 \tab 49 \tab 56 \tab 52 \tab 58 \tab 44 \tab 42 \tab 46 \tab 36 \tab 50 \tab 45 \tab 52 \tab 27 \tab 45 \tab 55 \tab 46 \tab 41 \tab 48 \tab 49 \tab 65 \tab 60 \tab 53 \tab 48 \tab 48 \tab 70 \tab 67 \tab 58 \tab 53 \tab 56 \tab 56 \tab 43 \tab 34 \tab 44 \tab 47 \tab 49 \tab 43 \tab 42 \tab 32 \tab 27 \tab 26 \tab 22 \tab 24 \tab 19 \tab 25 \tab 23 \tab 26 \tab 21 \tab 25 \tab 21 \tab 11 \tab 22 \tab 11 \tab 11 \tab 9 \tab 3 \tab 22 \tab 10 \tab 12 \tab - \tab - \tab 2867 \cr
   2018 \tab 22 \tab 26 \tab 15 \tab 27 \tab 40 \tab 29 \tab 38 \tab 43 \tab 31 \tab 39 \tab 45 \tab 43 \tab 50 \tab 34 \tab 43 \tab 42 \tab 65 \tab 40 \tab 40 \tab 43 \tab 38 \tab 55 \tab 41 \tab 40 \tab 40 \tab 39 \tab 40 \tab 29 \tab 42 \tab 30 \tab 33 \tab 31 \tab 37 \tab 41 \tab 29 \tab 48 \tab 35 \tab 48 \tab 50 \tab 39 \tab 37 \tab 46 \tab 46 \tab 39 \tab 24 \tab 37 \tab 33 \tab 39 \tab 36 \tab 27 \tab 39 \tab 41 \tab 45 \tab 33 \tab 22 \tab 20 \tab 29 \tab 23 \tab 19 \tab 28 \tab 11 \tab 9 \tab 12 \tab 9 \tab 12 \tab 7 \tab 10 \tab 12 \tab 5 \tab 8 \tab 29 \tab 7 \tab 14 \tab - \tab - \tab 2348 \cr
   2021 \tab 4 \tab 14 \tab 18 \tab 24 \tab 31 \tab 30 \tab 38 \tab 46 \tab 44 \tab 35 \tab 59 \tab 63 \tab 61 \tab 56 \tab 59 \tab 82 \tab 69 \tab 62 \tab 61 \tab 74 \tab 59 \tab 65 \tab 63 \tab 76 \tab 77 \tab 56 \tab 60 \tab 52 \tab 61 \tab 54 \tab 47 \tab 54 \tab 51 \tab 69 \tab 53 \tab 65 \tab 53 \tab 65 \tab 63 \tab 79 \tab 60 \tab 81 \tab 70 \tab 76 \tab 77 \tab 77 \tab 68 \tab 68 \tab 58 \tab 89 \tab 72 \tab 73 \tab 71 \tab 55 \tab 52 \tab 48 \tab 69 \tab 57 \tab 34 \tab 37 \tab 32 \tab 40 \tab 30 \tab 31 \tab 18 \tab 17 \tab 21 \tab 16 \tab 5 \tab 5 \tab 26 \tab 107 \tab 14 \tab - \tab 226 \tab 4032 \cr
   2022 \tab 22 \tab 29 \tab 48 \tab 46 \tab 46 \tab 53 \tab 45 \tab 45 \tab 58 \tab 46 \tab 57 \tab 61 \tab 60 \tab 68 \tab 76 \tab 69 \tab 61 \tab 58 \tab 69 \tab 47 \tab 57 \tab 72 \tab 62 \tab 48 \tab 67 \tab 45 \tab 55 \tab 56 \tab 52 \tab 53 \tab 46 \tab 51 \tab 40 \tab 55 \tab 47 \tab 38 \tab 50 \tab 60 \tab 56 \tab 53 \tab 44 \tab 51 \tab 75 \tab 57 \tab 60 \tab 48 \tab 50 \tab 65 \tab 67 \tab 44 \tab 68 \tab 56 \tab 67 \tab 40 \tab 54 \tab 34 \tab 43 \tab 34 \tab 31 \tab 32 \tab 26 \tab 30 \tab 22 \tab 15 \tab 15 \tab 13 \tab 21 \tab 8 \tab 5 \tab 4 \tab 19 \tab 208 \tab 11 \tab - \tab - \tab 3544 \cr
   Total \tab 267 \tab 904 \tab 951 \tab 1084 \tab 1159 \tab 1321 \tab 1296 \tab 1469 \tab 1438 \tab 1472 \tab 1548 \tab 1451 \tab 1571 \tab 1464 \tab 1566 \tab 1526 \tab 1552 \tab 1503 \tab 1492 \tab 1481 \tab 1466 \tab 1373 \tab 1412 \tab 1346 \tab 1345 \tab 1341 \tab 1289 \tab 1199 \tab 1236 \tab 1211 \tab 1193 \tab 1227 \tab 1148 \tab 1191 \tab 1128 \tab 1145 \tab 1106 \tab 1054 \tab 1137 \tab 1046 \tab 1078 \tab 1047 \tab 1085 \tab 975 \tab 1002 \tab 968 \tab 854 \tab 952 \tab 874 \tab 933 \tab 877 \tab 818 \tab 857 \tab 719 \tab 722 \tab 634 \tab 705 \tab 595 \tab 572 \tab 546 \tab 491 \tab 446 \tab 389 \tab 375 \tab 312 \tab 290 \tab 268 \tab 221 \tab 158 \tab 130 \tab 409 \tab 542 \tab 211 \tab 1 \tab 226 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{89} 89 or older
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
