% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_59.R
\name{aged}
\alias{aged}
\title{aged}
\source{
General Social Survey https://gss.norc.org
}
\description{
Should aged live with their children
}
\details{
Question 174. As you know, many older people share a home with their grown children. Do you think this is generally a good idea or a bad idea?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a bad idea \tab a good idea \tab depends \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 865 \tab 469 \tab 161 \tab 8 \tab 1 \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 795 \tab 460 \tab 225 \tab 9 \tab 1 \tab - \tab 1490 \cr
   1976 \tab - \tab 726 \tab 543 \tab 216 \tab 13 \tab 1 \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 741 \tab 531 \tab 254 \tab 6 \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 633 \tab 591 \tab 225 \tab 18 \tab 1 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 674 \tab 684 \tab 235 \tab 4 \tab 2 \tab - \tab 1599 \cr
   1984 \tab - \tab 531 \tab 734 \tab 194 \tab 11 \tab 3 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 686 \tab 592 \tab 182 \tab 10 \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 604 \tab 921 \tab 276 \tab 15 \tab 3 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 406 \tab 440 \tab 142 \tab 7 \tab 2 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 410 \tab 437 \tab 176 \tab 11 \tab 1 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 337 \tab 396 \tab 151 \tab 15 \tab - \tab - \tab 1372 \cr
   1991 \tab 500 \tab 384 \tab 451 \tab 167 \tab 9 \tab 6 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 369 \tab 455 \tab 218 \tab 13 \tab 2 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 721 \tab 925 \tab 339 \tab 20 \tab 6 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 632 \tab 883 \tab 372 \tab 29 \tab 9 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 658 \tab 853 \tab 371 \tab 23 \tab 6 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 592 \tab 962 \tab 313 \tab 21 \tab 8 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 301 \tab 433 \tab 176 \tab 2 \tab - \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 258 \tab 431 \tab 184 \tab 2 \tab 1 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 614 \tab 937 \tab 428 \tab 10 \tab - \tab - \tab 4510 \cr
   2008 \tab 658 \tab 456 \tab 660 \tab 239 \tab 8 \tab 2 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 392 \tab 716 \tab 255 \tab 10 \tab 4 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 425 \tab 684 \tab 212 \tab 16 \tab 1 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 490 \tab 881 \tab 302 \tab 13 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 912 \tab 534 \tab 1118 \tab 284 \tab 15 \tab 4 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 432 \tab 874 \tab 245 \tab 11 \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 2355 \tab 281 \tab 642 \tab 251 \tab 13 \tab 2 \tab - \tab 3544 \cr
   Total \tab 27505 \tab 14947 \tab 18703 \tab 6793 \tab 342 \tab 68 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} a good idea
\item \code{2} a bad idea
\item \code{3} depends
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
