% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_10.R
\name{agekdbrn}
\alias{agekdbrn}
\title{agekdbrn}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's age when 1st child born
}
\details{
Question 14. How old were you when your first child was born?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 45 \tab 9 \tab don't know \tab no answer \tab 44 \tab 46 \tab 48 \tab 49 \tab 65 \tab 43 \tab 47 \tab 11 \tab 50 \tab 51 \tab 54 \tab 53 \tab 56 \tab 52 \tab 55 \tab 57 \tab 10 \tab skipped on web \tab 90 \tab 97 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1377 \tab 1 \tab 2 \tab 5 \tab 8 \tab 39 \tab 63 \tab 109 \tab 125 \tab 142 \tab 128 \tab 101 \tab 106 \tab 114 \tab 117 \tab 92 \tab 91 \tab 82 \tab 61 \tab 57 \tab 39 \tab 28 \tab 18 \tab 17 \tab 14 \tab 10 \tab 3 \tab 6 \tab 6 \tab 6 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 19 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 824 \tab 1 \tab 5 \tab 7 \tab 19 \tab 58 \tab 85 \tab 133 \tab 151 \tab 182 \tab 203 \tab 170 \tab 135 \tab 130 \tab 139 \tab 121 \tab 103 \tab 82 \tab 53 \tab 64 \tab 39 \tab 32 \tab 28 \tab 19 \tab 26 \tab 12 \tab 13 \tab 11 \tab 2 \tab 6 \tab 4 \tab - \tab 2 \tab - \tab 10 \tab 28 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 802 \tab - \tab 5 \tab 7 \tab 20 \tab 48 \tab 99 \tab 124 \tab 176 \tab 166 \tab 177 \tab 144 \tab 156 \tab 111 \tab 151 \tab 92 \tab 89 \tab 93 \tab 66 \tab 70 \tab 36 \tab 52 \tab 22 \tab 19 \tab 18 \tab 15 \tab 9 \tab 5 \tab 9 \tab 3 \tab 7 \tab 6 \tab 1 \tab - \tab 3 \tab 28 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 813 \tab - \tab 1 \tab 6 \tab 20 \tab 53 \tab 97 \tab 142 \tab 154 \tab 164 \tab 164 \tab 162 \tab 118 \tab 130 \tab 140 \tab 97 \tab 96 \tab 78 \tab 74 \tab 64 \tab 37 \tab 38 \tab 31 \tab 21 \tab 24 \tab 12 \tab 8 \tab 10 \tab 8 \tab 4 \tab 4 \tab 4 \tab - \tab - \tab 7 \tab 32 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 799 \tab - \tab 4 \tab 3 \tab 15 \tab 46 \tab 79 \tab 122 \tab 158 \tab 180 \tab 187 \tab 156 \tab 134 \tab 127 \tab 108 \tab 94 \tab 96 \tab 78 \tab 62 \tab 69 \tab 45 \tab 47 \tab 39 \tab 18 \tab 25 \tab 21 \tab 15 \tab 5 \tab 5 \tab 4 \tab 3 \tab 3 \tab - \tab - \tab 2 \tab 9 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 767 \tab 1 \tab 2 \tab 6 \tab 12 \tab 48 \tab 72 \tab 132 \tab 168 \tab 177 \tab 181 \tab 157 \tab 115 \tab 131 \tab 114 \tab 104 \tab 119 \tab 86 \tab 56 \tab 87 \tab 52 \tab 55 \tab 33 \tab 27 \tab 26 \tab 12 \tab 12 \tab 10 \tab 11 \tab 6 \tab 3 \tab 2 \tab 1 \tab - \tab 10 \tab 9 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2348 \tab - \tab 2 \tab 8 \tab 22 \tab 62 \tab 89 \tab 134 \tab 156 \tab 201 \tab 216 \tab 148 \tab 149 \tab 120 \tab 132 \tab 122 \tab 91 \tab 80 \tab 84 \tab 79 \tab 42 \tab 49 \tab 31 \tab 21 \tab 21 \tab 23 \tab 15 \tab 13 \tab 6 \tab 10 \tab 4 \tab 5 \tab 1 \tab - \tab 1 \tab 13 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 521 \tab - \tab 2 \tab 5 \tab 14 \tab 40 \tab 64 \tab 102 \tab 104 \tab 135 \tab 149 \tab 115 \tab 92 \tab 79 \tab 93 \tab 88 \tab 71 \tab 67 \tab 46 \tab 57 \tab 26 \tab 30 \tab 19 \tab 19 \tab 26 \tab 13 \tab 7 \tab 11 \tab 5 \tab 2 \tab - \tab 4 \tab 1 \tab - \tab 6 \tab 7 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 562 \tab 1 \tab 4 \tab 10 \tab 16 \tab 44 \tab 70 \tab 117 \tab 110 \tab 125 \tab 133 \tab 95 \tab 87 \tab 88 \tab 69 \tab 64 \tab 70 \tab 65 \tab 48 \tab 52 \tab 42 \tab 33 \tab 23 \tab 15 \tab 16 \tab 19 \tab 13 \tab 6 \tab 6 \tab 6 \tab 6 \tab 6 \tab 2 \tab - \tab 2 \tab 10 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 536 \tab - \tab 1 \tab 7 \tab 20 \tab 30 \tab 68 \tab 101 \tab 102 \tab 110 \tab 134 \tab 78 \tab 95 \tab 82 \tab 94 \tab 70 \tab 78 \tab 58 \tab 52 \tab 55 \tab 32 \tab 32 \tab 28 \tab 29 \tab 21 \tab 13 \tab 11 \tab 10 \tab 7 \tab 6 \tab 3 \tab 1 \tab - \tab - \tab 3 \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 704 \tab 2 \tab 2 \tab 4 \tab 16 \tab 63 \tab 80 \tab 117 \tab 135 \tab 134 \tab 168 \tab 120 \tab 114 \tab 107 \tab 123 \tab 70 \tab 62 \tab 63 \tab 60 \tab 86 \tab 37 \tab 51 \tab 38 \tab 27 \tab 40 \tab 21 \tab 16 \tab 14 \tab 10 \tab 11 \tab 6 \tab 5 \tab 2 \tab - \tab 7 \tab 10 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 797 \tab 1 \tab 4 \tab 12 \tab 16 \tab 67 \tab 84 \tab 135 \tab 165 \tab 164 \tab 170 \tab 115 \tab 124 \tab 101 \tab 144 \tab 109 \tab 89 \tab 76 \tab 71 \tab 97 \tab 50 \tab 55 \tab 41 \tab 31 \tab 38 \tab 19 \tab 18 \tab 13 \tab 13 \tab 15 \tab 4 \tab 5 \tab 3 \tab 1 \tab 3 \tab 11 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 663 \tab 1 \tab 2 \tab 6 \tab 20 \tab 35 \tab 70 \tab 96 \tab 131 \tab 126 \tab 142 \tab 113 \tab 101 \tab 115 \tab 120 \tab 63 \tab 67 \tab 87 \tab 59 \tab 83 \tab 37 \tab 43 \tab 38 \tab 20 \tab 28 \tab 9 \tab 10 \tab 10 \tab 7 \tab 10 \tab 4 \tab 3 \tab 1 \tab - \tab 10 \tab 9 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1212 \tab - \tab 4 \tab 5 \tab 21 \tab 55 \tab 89 \tab 132 \tab 157 \tab 213 \tab 207 \tab 158 \tab 170 \tab 171 \tab 177 \tab 134 \tab 143 \tab 144 \tab 109 \tab 147 \tab 72 \tab 94 \tab 80 \tab 70 \tab 58 \tab 41 \tab 29 \tab 37 \tab 29 \tab 20 \tab 9 \tab 5 \tab 5 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 17 \tab - \tab - \tab 4032 \cr
   2022 \tab 1107 \tab 1 \tab 2 \tab 7 \tab 24 \tab 53 \tab 97 \tab 133 \tab 168 \tab 176 \tab 207 \tab 147 \tab 131 \tab 129 \tab 145 \tab 138 \tab 122 \tab 117 \tab 91 \tab 121 \tab 46 \tab 59 \tab 52 \tab 42 \tab 45 \tab 28 \tab 24 \tab 24 \tab 22 \tab 14 \tab 9 \tab 11 \tab 2 \tab 1 \tab 1 \tab 8 \tab 2 \tab 3 \tab - \tab 2 \tab - \tab 6 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 22 \tab 1 \tab 1 \tab 3544 \cr
   Total \tab 43220 \tab 9 \tab 42 \tab 98 \tab 263 \tab 741 \tab 1206 \tab 1829 \tab 2160 \tab 2395 \tab 2566 \tab 1979 \tab 1827 \tab 1735 \tab 1866 \tab 1458 \tab 1387 \tab 1256 \tab 992 \tab 1188 \tab 632 \tab 698 \tab 521 \tab 395 \tab 426 \tab 268 \tab 203 \tab 185 \tab 146 \tab 123 \tab 67 \tab 62 \tab 22 \tab 4 \tab 66 \tab 198 \tab 17 \tab 16 \tab 6 \tab 7 \tab 1 \tab 26 \tab 14 \tab 3 \tab 9 \tab 4 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 39 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
