% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_03.R
\name{agewed}
\alias{agewed}
\title{agewed}
\source{
General Social Survey https://gss.norc.org
}
\description{
Age when first married
}
\details{
Question 4a. If ever married: How old were you when you first married?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 42 \tab 43 \tab 45 \tab 46 \tab 48 \tab 52 \tab 55 \tab 58 \tab iap \tab no answer \tab 41 \tab 44 \tab 63 \tab 49 \tab 51 \tab 60 \tab 12 \tab 50 \tab 73 \tab don't know \tab 47 \tab 56 \tab 62 \tab 68 \tab 54 \tab 61 \tab 53 \tab 57 \tab 59 \tab 65 \tab 70 \tab 90 \tab Total \cr
   1972 \tab 4 \tab 8 \tab 18 \tab 41 \tab 81 \tab 146 \tab 131 \tab 153 \tab 172 \tab 136 \tab 105 \tab 80 \tab 73 \tab 47 \tab 42 \tab 42 \tab 22 \tab 17 \tab 14 \tab 7 \tab 9 \tab 6 \tab 10 \tab 4 \tab 3 \tab 5 \tab 5 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 210 \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 2 \tab 6 \tab 15 \tab 37 \tab 87 \tab 136 \tab 138 \tab 120 \tab 176 \tab 114 \tab 101 \tab 79 \tab 61 \tab 54 \tab 31 \tab 29 \tab 31 \tab 25 \tab 8 \tab 13 \tab 7 \tab 4 \tab 8 \tab 2 \tab 5 \tab - \tab 3 \tab 6 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 193 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1 \tab 6 \tab 15 \tab 43 \tab 66 \tab 129 \tab 149 \tab 138 \tab 187 \tab 111 \tab 93 \tab 73 \tab 74 \tab 44 \tab 28 \tab 21 \tab 21 \tab 26 \tab 10 \tab 13 \tab 12 \tab 8 \tab 5 \tab 4 \tab 2 \tab 4 \tab 2 \tab 4 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 184 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 2 \tab 6 \tab 15 \tab 55 \tab 72 \tab 130 \tab 134 \tab 132 \tab 173 \tab 109 \tab 85 \tab 68 \tab 78 \tab 37 \tab 35 \tab 25 \tab 22 \tab 17 \tab 14 \tab 14 \tab 4 \tab 5 \tab 10 \tab 5 \tab 4 \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 211 \tab 9 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 5 \tab 11 \tab 12 \tab 33 \tab 76 \tab 135 \tab 132 \tab 117 \tab 156 \tab 120 \tab 112 \tab 80 \tab 70 \tab 52 \tab 47 \tab 25 \tab 23 \tab 25 \tab 7 \tab 9 \tab 8 \tab 8 \tab 6 \tab 2 \tab 1 \tab 2 \tab 2 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 207 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1 \tab 6 \tab 11 \tab 56 \tab 71 \tab 126 \tab 135 \tab 127 \tab 161 \tab 129 \tab 89 \tab 73 \tab 68 \tab 47 \tab 51 \tab 32 \tab 24 \tab 19 \tab 9 \tab 17 \tab 7 \tab 7 \tab 6 \tab 2 \tab 4 \tab 4 \tab 1 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 225 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 4 \tab 4 \tab 16 \tab 43 \tab 77 \tab 132 \tab 171 \tab 135 \tab 160 \tab 107 \tab 97 \tab 88 \tab 62 \tab 39 \tab 30 \tab 25 \tab 19 \tab 21 \tab 10 \tab 9 \tab 7 \tab 5 \tab 6 \tab 4 \tab 2 \tab 1 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 237 \tab 7 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1 \tab 5 \tab 16 \tab 41 \tab 81 \tab 135 \tab 143 \tab 130 \tab 154 \tab 88 \tab 92 \tab 93 \tab 58 \tab 49 \tab 28 \tab 16 \tab 17 \tab 13 \tab 7 \tab 10 \tab 4 \tab 7 \tab 6 \tab 3 \tab 1 \tab 2 \tab 3 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 242 \tab 11 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1 \tab 7 \tab 27 \tab 48 \tab 86 \tab 167 \tab 184 \tab 150 \tab 183 \tab 133 \tab 101 \tab 83 \tab 75 \tab 59 \tab 50 \tab 24 \tab 23 \tab 27 \tab 12 \tab 7 \tab 9 \tab 6 \tab 7 \tab 5 \tab 4 \tab 4 \tab 2 \tab 4 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 351 \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 2 \tab 4 \tab 13 \tab 43 \tab 82 \tab 128 \tab 145 \tab 137 \tab 176 \tab 112 \tab 103 \tab 89 \tab 85 \tab 44 \tab 37 \tab 30 \tab 15 \tab 22 \tab 7 \tab 14 \tab 9 \tab 3 \tab 3 \tab 2 \tab - \tab 5 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 265 \tab 8 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 1 \tab 13 \tab 47 \tab 70 \tab 116 \tab 140 \tab 130 \tab 147 \tab 117 \tab 69 \tab 78 \tab 56 \tab 41 \tab 26 \tab 31 \tab 17 \tab 13 \tab 11 \tab 8 \tab 8 \tab 6 \tab 7 \tab 5 \tab 3 \tab 3 \tab 2 \tab 3 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 282 \tab 15 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 3 \tab 12 \tab 41 \tab 68 \tab 138 \tab 150 \tab 137 \tab 163 \tab 104 \tab 93 \tab 77 \tab 63 \tab 48 \tab 52 \tab 29 \tab 13 \tab 21 \tab 12 \tab 5 \tab 4 \tab 4 \tab 7 \tab 3 \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 268 \tab 6 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1 \tab 3 \tab 10 \tab 35 \tab 74 \tab 132 \tab 133 \tab 117 \tab 143 \tab 128 \tab 81 \tab 70 \tab 76 \tab 49 \tab 34 \tab 26 \tab 18 \tab 13 \tab 10 \tab 9 \tab 4 \tab 8 \tab 5 \tab 2 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 266 \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 6 \tab 7 \tab 19 \tab 45 \tab 70 \tab 161 \tab 156 \tab 140 \tab 180 \tab 108 \tab 113 \tab 85 \tab 87 \tab 48 \tab 44 \tab 36 \tab 21 \tab 22 \tab 12 \tab 8 \tab 8 \tab - \tab 8 \tab 4 \tab 1 \tab 4 \tab 6 \tab - \tab 6 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 380 \tab 21 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1 \tab 4 \tab 7 \tab 38 \tab 70 \tab 139 \tab 138 \tab 127 \tab 144 \tab 92 \tab 83 \tab 66 \tab 63 \tab 45 \tab 49 \tab 28 \tab 9 \tab 19 \tab 6 \tab 11 \tab 5 \tab 2 \tab 6 \tab 3 \tab 6 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 298 \tab 11 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1 \tab 4 \tab 12 \tab 37 \tab 63 \tab 121 \tab 136 \tab 136 \tab 162 \tab 87 \tab 116 \tab 72 \tab 64 \tab 32 \tab 45 \tab 28 \tab 28 \tab 27 \tab 9 \tab 14 \tab 3 \tab 3 \tab 3 \tab 4 \tab 3 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 301 \tab 13 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 5 \tab 8 \tab 26 \tab 61 \tab 94 \tab 107 \tab 122 \tab 132 \tab 102 \tab 79 \tab 61 \tab 68 \tab 48 \tab 47 \tab 30 \tab 20 \tab 25 \tab 11 \tab 11 \tab 9 \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 265 \tab 17 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 2 \tab 9 \tab 12 \tab 43 \tab 53 \tab 133 \tab 123 \tab 105 \tab 140 \tab 97 \tab 97 \tab 74 \tab 66 \tab 38 \tab 43 \tab 31 \tab 20 \tab 18 \tab 11 \tab 13 \tab 8 \tab 11 \tab 5 \tab 2 \tab 6 \tab 2 \tab 4 \tab 4 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 324 \tab 12 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1 \tab 4 \tab 8 \tab 36 \tab 47 \tab 126 \tab 137 \tab 129 \tab 145 \tab 106 \tab 90 \tab 88 \tab 79 \tab 63 \tab 54 \tab 29 \tab 36 \tab 27 \tab 18 \tab 21 \tab 11 \tab 5 \tab 7 \tab 7 \tab 3 \tab 4 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 300 \tab 12 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1 \tab - \tab 9 \tab 30 \tab 52 \tab 126 \tab 106 \tab 126 \tab 131 \tab 96 \tab 98 \tab 84 \tab 80 \tab 40 \tab 41 \tab 34 \tab 30 \tab 32 \tab 17 \tab 10 \tab 12 \tab 10 \tab 6 \tab 4 \tab - \tab 2 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1790 \tab 10 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2 \tab 3 \tab 4 \tab 25 \tab 46 \tab 117 \tab 98 \tab 111 \tab 139 \tab 95 \tab 83 \tab 72 \tab 71 \tab 54 \tab 53 \tab 32 \tab 23 \tab 25 \tab 15 \tab 15 \tab 12 \tab 11 \tab 11 \tab 3 \tab 5 \tab 5 \tab 4 \tab 9 \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 3342 \tab 6 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 38 \tab 106 \tab 272 \tab 843 \tab 1453 \tab 2767 \tab 2886 \tab 2719 \tab 3324 \tab 2291 \tab 1980 \tab 1633 \tab 1477 \tab 978 \tab 867 \tab 603 \tab 452 \tab 454 \tab 230 \tab 238 \tab 160 \tab 119 \tab 135 \tab 72 \tab 58 \tab 62 \tab 56 \tab 62 \tab 35 \tab 22 \tab 23 \tab 9 \tab 7 \tab 9 \tab 8 \tab 5 \tab 45641 \tab 187 \tab 18 \tab 12 \tab 2 \tab 8 \tab 3 \tab 4 \tab 4 \tab 8 \tab 2 \tab 19 \tab 7 \tab 4 \tab 1 \tab 1 \tab 6 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
