% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_60.R
\name{anomia6}
\alias{anomia6}
\title{Not fair to bring child into world}
\source{
General Social Survey https://gss.norc.org
}
\description{
anomia6
}
\details{
Question 178. Now I'm going to read you several more statements. Some people agree with a statement, others disagree. As I read each one, tell me whether you more or less agree with it, or more or less disagree. f. It's hardly fair to bring a child into the world with the way things look for the future.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 544 \tab 924 \tab 35 \tab 1 \tab 1504 \cr
   1974 \tab - \tab 531 \tab 913 \tab 38 \tab 2 \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 621 \tab 821 \tab 53 \tab 4 \tab 1499 \cr
   1977 \tab - \tab 576 \tab 905 \tab 44 \tab 5 \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 685 \tab 743 \tab 38 \tab 2 \tab 1468 \cr
   1982 \tab - \tab 657 \tab 1101 \tab 72 \tab 30 \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 587 \tab 861 \tab 22 \tab 3 \tab 1473 \cr
   1985 \tab - \tab 496 \tab 990 \tab 44 \tab 4 \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 768 \tab 995 \tab 52 \tab 4 \tab 1819 \cr
   1988 \tab 504 \tab 371 \tab 580 \tab 24 \tab 2 \tab 1481 \cr
   1989 \tab 504 \tab 385 \tab 620 \tab 24 \tab 4 \tab 1537 \cr
   1990 \tab 455 \tab 320 \tab 566 \tab 27 \tab 4 \tab 1372 \cr
   1991 \tab 524 \tab 346 \tab 614 \tab 31 \tab 2 \tab 1517 \cr
   1993 \tab 531 \tab 442 \tab 600 \tab 30 \tab 3 \tab 1606 \cr
   1994 \tab 1737 \tab 554 \tab 670 \tab 30 \tab 1 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 51969 \tab 7883 \tab 11903 \tab 564 \tab 71 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} agree
\item \code{2} disagree
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
