% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_15.R
\name{babies}
\alias{babies}
\title{Household members less than 6 yrs old}
\source{
General Social Survey https://gss.norc.org
}
\description{
babies
}
\details{
Question 34a. Number of members under 6 years of age.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab 0 household members under 6 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab no answer \tab 6 \tab not available in this release \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 1215 \tab 228 \tab 129 \tab 34 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1150 \tab 216 \tab 110 \tab 22 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1127 \tab 227 \tab 101 \tab 24 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1167 \tab 228 \tab 83 \tab 10 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1160 \tab 224 \tab 88 \tab 21 \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1248 \tab 197 \tab 62 \tab 13 \tab 1 \tab - \tab 9 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1232 \tab 195 \tab 78 \tab 11 \tab 2 \tab - \tab 14 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1179 \tab 200 \tab 66 \tab 10 \tab 2 \tab - \tab 11 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1507 \tab 240 \tab 90 \tab 18 \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1275 \tab 212 \tab 90 \tab 14 \tab 3 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1203 \tab 176 \tab 74 \tab 13 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1244 \tab 200 \tab 66 \tab 17 \tab 2 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1189 \tab 194 \tab 70 \tab 10 \tab 2 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1509 \tab 215 \tab 76 \tab 16 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1200 \tab 202 \tab 58 \tab 10 \tab - \tab - \tab 11 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1233 \tab 192 \tab 77 \tab 13 \tab 1 \tab - \tab 21 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1134 \tab 157 \tab 67 \tab 4 \tab - \tab - \tab 10 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1231 \tab 183 \tab 87 \tab 8 \tab 1 \tab 1 \tab 6 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1321 \tab 198 \tab 75 \tab 9 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2485 \tab 369 \tab 108 \tab 23 \tab 4 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2408 \tab 340 \tab 132 \tab 22 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2401 \tab 308 \tab 102 \tab 15 \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2322 \tab 315 \tab 115 \tab 16 \tab 2 \tab - \tab 47 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2379 \tab 259 \tab 74 \tab 22 \tab 2 \tab - \tab 29 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2439 \tab 241 \tab 96 \tab 15 \tab 1 \tab - \tab 20 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 3850 \tab 421 \tab 167 \tab 29 \tab 5 \tab - \tab 38 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1697 \tab 187 \tab 93 \tab 15 \tab 2 \tab - \tab 29 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1759 \tab 201 \tab 60 \tab 11 \tab 3 \tab 2 \tab 8 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1686 \tab 180 \tab 75 \tab 12 \tab 1 \tab - \tab 20 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2275 \tab 177 \tab 68 \tab 11 \tab 3 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2550 \tab 192 \tab 86 \tab 13 \tab 2 \tab - \tab 24 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2073 \tab 185 \tab 53 \tab 7 \tab - \tab 1 \tab 29 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab 4032 \cr
   2022 \tab 1577 \tab 130 \tab 56 \tab 4 \tab 1 \tab - \tab 8 \tab - \tab - \tab 1764 \tab 4 \tab 3544 \cr
   Total \tab 55425 \tab 7389 \tab 2832 \tab 492 \tab 62 \tab 9 \tab 380 \tab 1 \tab 4032 \tab 1764 \tab 4 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} 0 household members under 6
\item \code{8} 8 or more
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
