% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{ballot}
\alias{ballot}
\title{Ballot used for interview}
\source{
General Social Survey https://gss.norc.org
}
\description{
ballot
}
\details{
Question 1659. Ballot Administered.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab ballot a \tab ballot b \tab ballot c \tab ballot d \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 484 \tab 504 \tab 493 \tab - \tab 1481 \cr
   1989 \tab - \tab 502 \tab 504 \tab 531 \tab - \tab 1537 \cr
   1990 \tab - \tab 473 \tab 455 \tab 444 \tab - \tab 1372 \cr
   1991 \tab - \tab 500 \tab 524 \tab 493 \tab - \tab 1517 \cr
   1993 \tab - \tab 549 \tab 531 \tab 526 \tab - \tab 1606 \cr
   1994 \tab - \tab 981 \tab 996 \tab 1015 \tab - \tab 2992 \cr
   1996 \tab - \tab 979 \tab 981 \tab 944 \tab - \tab 2904 \cr
   1998 \tab - \tab 921 \tab 950 \tab 961 \tab - \tab 2832 \cr
   2000 \tab - \tab 921 \tab 956 \tab 940 \tab - \tab 2817 \cr
   2002 \tab - \tab 933 \tab 911 \tab 921 \tab - \tab 2765 \cr
   2004 \tab - \tab 952 \tab 941 \tab 919 \tab - \tab 2812 \cr
   2006 \tab - \tab 1003 \tab 989 \tab 1000 \tab 1518 \tab 4510 \cr
   2008 \tab - \tab 658 \tab 671 \tab 694 \tab - \tab 2023 \cr
   2010 \tab - \tab 667 \tab 763 \tab 614 \tab - \tab 2044 \cr
   2012 \tab - \tab 636 \tab 666 \tab 672 \tab - \tab 1974 \cr
   2014 \tab - \tab 851 \tab 824 \tab 863 \tab - \tab 2538 \cr
   2016 \tab - \tab 911 \tab 977 \tab 979 \tab - \tab 2867 \cr
   2018 \tab - \tab 785 \tab 774 \tab 789 \tab - \tab 2348 \cr
   2021 \tab - \tab 1360 \tab 1357 \tab 1315 \tab - \tab 4032 \cr
   2022 \tab - \tab 1173 \tab 1203 \tab 1168 \tab - \tab 3544 \cr
   Total \tab 21875 \tab 16239 \tab 16477 \tab 16281 \tab 1518 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} ballot a
\item \code{2} ballot b
\item \code{3} ballot c
\item \code{4} ballot d
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
