% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_254.R
\name{betrlang}
\alias{betrlang}
\title{betrlang}
\source{
General Social Survey https://gss.norc.org
}
\description{
Which language r speaks more fluent
}
\details{
Question 746b. Which language do you speak more fluent?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab language 1 \tab language 2 \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 3837 \tab 635 \tab 36 \tab 2 \tab - \tab - \tab 4510 \cr
   2008 \tab 1471 \tab 515 \tab 32 \tab - \tab 5 \tab - \tab 2023 \cr
   2010 \tab 1508 \tab 500 \tab 33 \tab - \tab 3 \tab - \tab 2044 \cr
   2012 \tab 1401 \tab 537 \tab 33 \tab 1 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 1790 \tab 699 \tab 43 \tab - \tab 6 \tab - \tab 2538 \cr
   2016 \tab 2045 \tab 749 \tab 62 \tab 8 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 1623 \tab 690 \tab 31 \tab 1 \tab 3 \tab - \tab 2348 \cr
   2021 \tab 2972 \tab 1004 \tab 46 \tab 1 \tab 2 \tab 7 \tab 4032 \cr
   2022 \tab 2605 \tab 887 \tab 49 \tab - \tab 1 \tab 2 \tab 3544 \cr
   Total \tab 65762 \tab 6216 \tab 365 \tab 13 \tab 25 \tab 9 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} language 1
\item \code{2} language 2
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
