% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_14.R
\name{born}
\alias{born}
\title{born}
\source{
General Social Survey https://gss.norc.org
}
\description{
Was r born in this country
}
\details{
Question 31. Were you born in this country?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 105 \tab 2 \tab 1423 \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 92 \tab 2 \tab 1438 \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 104 \tab 2 \tab 1362 \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 101 \tab 3 \tab 1755 \tab 1 \tab - \tab 1860 \cr
   1983 \tab - \tab 113 \tab 1 \tab 1485 \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 88 \tab 8 \tab 1377 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 95 \tab 3 \tab 1436 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 108 \tab 8 \tab 1354 \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 103 \tab 10 \tab 1706 \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 95 \tab - \tab 1386 \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 102 \tab 6 \tab 1429 \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 89 \tab 11 \tab 1272 \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 93 \tab 10 \tab 1413 \tab 1 \tab - \tab 1517 \cr
   1993 \tab - \tab 117 \tab 8 \tab 1481 \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 214 \tab 29 \tab 2749 \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 238 \tab 11 \tab 2655 \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 242 \tab 12 \tab 2577 \tab 1 \tab - \tab 2832 \cr
   2000 \tab - \tab 271 \tab 7 \tab 2537 \tab 2 \tab - \tab 2817 \cr
   2002 \tab - \tab 256 \tab 3 \tab 2503 \tab 3 \tab - \tab 2765 \cr
   2004 \tab - \tab 272 \tab 1 \tab 2539 \tab - \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 417 \tab - \tab 2575 \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 265 \tab - \tab 1758 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 248 \tab 1 \tab 1795 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 271 \tab 3 \tab 1700 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 354 \tab - \tab 2184 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 355 \tab 5 \tab 2507 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 302 \tab 1 \tab 2045 \tab - \tab - \tab 2348 \cr
   2021 \tab 47 \tab 444 \tab - \tab 3516 \tab 16 \tab 9 \tab 4032 \cr
   2022 \tab 5 \tab 445 \tab 2 \tab 3077 \tab 7 \tab 8 \tab 3544 \cr
   Total \tab 9160 \tab 5999 \tab 149 \tab 57034 \tab 31 \tab 17 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
