% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_600.R
\name{canada}
\alias{canada}
\title{canada}
\source{
General Social Survey https://gss.norc.org
}
\description{
Liking for canada
}
\details{
Question 103. You will notice that the boxes on this card go from the highest position of "plus 5" for a country which you like very much, to the lowest position of "minus 5" for a country you dislike very much. How far up the scale or how far down the scale would you rate the following countries? d. Canada
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab not available in this year \tab +1 \tab +2 \tab +3 \tab +4 \tab +5 \tab -1 \tab -2 \tab -3 \tab -4 \tab -5 \tab dk, na, iap \tab iap \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 94 \tab 111 \tab 210 \tab 331 \tab 621 \tab 20 \tab 8 \tab 13 \tab 4 \tab 14 \tab 58 \tab - \tab 1484 \cr
   1975 \tab - \tab 93 \tab 116 \tab 271 \tab 341 \tab 535 \tab 28 \tab 9 \tab 13 \tab 4 \tab 6 \tab 74 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 98 \tab 119 \tab 269 \tab 331 \tab 552 \tab 27 \tab 12 \tab 11 \tab 3 \tab 14 \tab 94 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 122 \tab 161 \tab 280 \tab 389 \tab 718 \tab 33 \tab 17 \tab 7 \tab 7 \tab 12 \tab 114 \tab - \tab 1860 \cr
   1983 \tab - \tab 105 \tab 138 \tab 293 \tab 369 \tab 581 \tab 19 \tab 9 \tab 9 \tab 2 \tab 7 \tab 67 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 77 \tab 132 \tab 287 \tab 334 \tab 595 \tab 17 \tab 5 \tab 9 \tab 7 \tab 6 \tab 65 \tab - \tab 1534 \cr
   1986 \tab - \tab 114 \tab 124 \tab 244 \tab 305 \tab 611 \tab 10 \tab 7 \tab 8 \tab 4 \tab 9 \tab 34 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 72 \tab 81 \tab 136 \tab 197 \tab 428 \tab 12 \tab 5 \tab 7 \tab 2 \tab 6 \tab 535 \tab - \tab 1481 \cr
   1989 \tab - \tab 81 \tab 65 \tab 166 \tab 212 \tab 373 \tab 22 \tab 18 \tab 3 \tab 5 \tab 12 \tab 580 \tab - \tab 1537 \cr
   1990 \tab - \tab 95 \tab 81 \tab 160 \tab 178 \tab 334 \tab 8 \tab 8 \tab 3 \tab 4 \tab 9 \tab 492 \tab - \tab 1372 \cr
   1991 \tab - \tab 74 \tab 76 \tab 161 \tab 210 \tab 414 \tab 14 \tab 6 \tab 5 \tab 2 \tab 13 \tab 542 \tab - \tab 1517 \cr
   1993 \tab - \tab 122 \tab 102 \tab 230 \tab 202 \tab 346 \tab 7 \tab 12 \tab 6 \tab 4 \tab 8 \tab 567 \tab - \tab 1606 \cr
   1994 \tab - \tab 49 \tab 47 \tab 102 \tab 103 \tab 164 \tab 7 \tab 4 \tab 4 \tab 2 \tab 2 \tab 2508 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 47374 \tab 1196 \tab 1353 \tab 2809 \tab 3502 \tab 6272 \tab 224 \tab 120 \tab 98 \tab 50 \tab 118 \tab 5730 \tab 3544 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} +5
\item \code{1} +4
\item \code{2} +3
\item \code{3} +2
\item \code{4} +1
\item \code{5} -1
\item \code{6} -2
\item \code{7} -3
\item \code{8} -4
\item \code{9} -5
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
