% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_32.R
\name{cappun}
\alias{cappun}
\title{Favor or oppose death penalty for murder}
\source{
General Social Survey https://gss.norc.org
}
\description{
cappun
}
\details{
Question 82b. Do you favor or oppose the death enalty for persons convicted of murder?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab favor \tab no answer \tab oppose \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 76 \tab 933 \tab 4 \tab 471 \tab - \tab 1484 \cr
   1975 \tab - \tab 100 \tab 891 \tab 7 \tab 492 \tab - \tab 1490 \cr
   1976 \tab - \tab 70 \tab 980 \tab 3 \tab 446 \tab - \tab 1499 \cr
   1977 \tab - \tab 97 \tab 1022 \tab 10 \tab 401 \tab - \tab 1530 \cr
   1978 \tab - \tab 89 \tab 1016 \tab - \tab 427 \tab - \tab 1532 \cr
   1980 \tab - \tab 89 \tab 982 \tab 7 \tab 390 \tab - \tab 1468 \cr
   1982 \tab - \tab 118 \tab 1263 \tab 5 \tab 474 \tab - \tab 1860 \cr
   1983 \tab - \tab 74 \tab 1169 \tab 2 \tab 354 \tab - \tab 1599 \cr
   1984 \tab - \tab 86 \tab 1029 \tab 11 \tab 347 \tab - \tab 1473 \cr
   1985 \tab - \tab 75 \tab 1154 \tab 8 \tab 297 \tab - \tab 1534 \cr
   1986 \tab - \tab 76 \tab 1046 \tab 4 \tab 344 \tab - \tab 1470 \cr
   1987 \tab - \tab 121 \tab 1166 \tab 13 \tab 519 \tab - \tab 1819 \cr
   1988 \tab - \tab 102 \tab 1045 \tab 6 \tab 328 \tab - \tab 1481 \cr
   1989 \tab - \tab 85 \tab 1133 \tab 5 \tab 314 \tab - \tab 1537 \cr
   1990 \tab - \tab 84 \tab 1018 \tab 6 \tab 264 \tab - \tab 1372 \cr
   1991 \tab - \tab 93 \tab 1078 \tab 10 \tab 336 \tab - \tab 1517 \cr
   1993 \tab - \tab 112 \tab 1151 \tab 6 \tab 337 \tab - \tab 1606 \cr
   1994 \tab - \tab 184 \tab 2215 \tab 13 \tab 580 \tab - \tab 2992 \cr
   1996 \tab - \tab 214 \tab 2059 \tab 10 \tab 621 \tab - \tab 2904 \cr
   1998 \tab - \tab 220 \tab 1906 \tab 13 \tab 693 \tab - \tab 2832 \cr
   2000 \tab - \tab 228 \tab 1764 \tab 24 \tab 801 \tab - \tab 2817 \cr
   2002 \tab 1393 \tab 57 \tab 899 \tab 7 \tab 409 \tab - \tab 2765 \cr
   2004 \tab 1472 \tab 68 \tab 855 \tab 15 \tab 402 \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 155 \tab 1885 \tab 22 \tab 930 \tab - \tab 4510 \cr
   2008 \tab - \tab 105 \tab 1263 \tab 16 \tab 639 \tab - \tab 2023 \cr
   2010 \tab - \tab 101 \tab 1297 \tab 22 \tab 624 \tab - \tab 2044 \cr
   2012 \tab - \tab 127 \tab 1183 \tab 23 \tab 641 \tab - \tab 1974 \cr
   2014 \tab - \tab 139 \tab 1530 \tab 20 \tab 849 \tab - \tab 2538 \cr
   2016 \tab - \tab 140 \tab 1629 \tab 32 \tab 1066 \tab - \tab 2867 \cr
   2018 \tab - \tab 125 \tab 1385 \tab 30 \tab 808 \tab - \tab 2348 \cr
   2021 \tab - \tab 20 \tab 2222 \tab 1 \tab 1735 \tab 54 \tab 4032 \cr
   2022 \tab - \tab 155 \tab 2013 \tab 33 \tab 1327 \tab 16 \tab 3544 \cr
   Total \tab 7500 \tab 3585 \tab 42181 \tab 388 \tab 18666 \tab 70 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} favor
\item \code{2} oppose
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
