% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_09.R
\name{childs}
\alias{childs}
\title{Number of children}
\source{
General Social Survey https://gss.norc.org
}
\description{
childs
}
\details{
Question 12. How many children have you ever had? Please count all that were born alive at any time (including any you had from a previous marriage).
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 or more \tab don't know \tab iap \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 373 \tab 240 \tab 340 \tab 276 \tab 169 \tab 87 \tab 52 \tab 28 \tab 40 \tab 8 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 378 \tab 200 \tab 357 \tab 233 \tab 157 \tab 65 \tab 45 \tab 22 \tab 43 \tab 4 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 332 \tab 226 \tab 319 \tab 245 \tab 141 \tab 97 \tab 52 \tab 25 \tab 40 \tab 7 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 386 \tab 242 \tab 330 \tab 234 \tab 132 \tab 64 \tab 36 \tab 19 \tab 42 \tab 5 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 392 \tab 244 \tab 328 \tab 227 \tab 145 \tab 65 \tab 28 \tab 32 \tab 36 \tab 2 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 383 \tab 256 \tab 356 \tab 229 \tab 130 \tab 70 \tab 38 \tab 19 \tab 35 \tab 14 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 417 \tab 240 \tab 366 \tab 222 \tab 134 \tab 63 \tab 39 \tab 24 \tab 21 \tab 6 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 389 \tab 236 \tab 359 \tab 213 \tab 120 \tab 64 \tab 36 \tab 18 \tab 30 \tab 3 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 497 \tab 326 \tab 423 \tab 259 \tab 145 \tab 82 \tab 40 \tab 27 \tab 58 \tab 3 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 431 \tab 261 \tab 375 \tab 259 \tab 116 \tab 61 \tab 44 \tab 25 \tab 24 \tab 3 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 431 \tab 230 \tab 328 \tab 212 \tab 134 \tab 55 \tab 17 \tab 19 \tab 30 \tab 17 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 430 \tab 226 \tab 349 \tab 241 \tab 132 \tab 76 \tab 28 \tab 17 \tab 31 \tab 4 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 399 \tab 214 \tab 357 \tab 218 \tab 128 \tab 61 \tab 49 \tab 14 \tab 28 \tab 2 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 517 \tab 289 \tab 395 \tab 291 \tab 146 \tab 64 \tab 35 \tab 21 \tab 54 \tab 7 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 395 \tab 239 \tab 358 \tab 208 \tab 135 \tab 71 \tab 38 \tab 13 \tab 23 \tab 1 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 427 \tab 230 \tab 397 \tab 236 \tab 103 \tab 70 \tab 30 \tab 20 \tab 22 \tab 2 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 392 \tab 208 \tab 341 \tab 215 \tab 117 \tab 50 \tab 26 \tab 14 \tab 7 \tab 2 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 419 \tab 255 \tab 375 \tab 215 \tab 127 \tab 54 \tab 24 \tab 23 \tab 17 \tab 8 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 447 \tab 259 \tab 423 \tab 243 \tab 122 \tab 64 \tab 14 \tab 8 \tab 21 \tab 5 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 820 \tab 484 \tab 795 \tab 447 \tab 239 \tab 103 \tab 39 \tab 23 \tab 33 \tab 9 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 822 \tab 461 \tab 770 \tab 420 \tab 222 \tab 94 \tab 48 \tab 27 \tab 25 \tab 15 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 802 \tab 474 \tab 743 \tab 411 \tab 209 \tab 86 \tab 47 \tab 19 \tab 34 \tab 7 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 813 \tab 455 \tab 719 \tab 449 \tab 188 \tab 93 \tab 32 \tab 25 \tab 27 \tab 16 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 799 \tab 469 \tab 657 \tab 481 \tab 185 \tab 73 \tab 40 \tab 22 \tab 34 \tab 5 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 767 \tab 474 \tab 730 \tab 459 \tab 221 \tab 79 \tab 39 \tab 18 \tab 21 \tab 4 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 1216 \tab 710 \tab 1147 \tab 738 \tab 386 \tab 139 \tab 83 \tab 43 \tab 35 \tab 13 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 521 \tab 323 \tab 524 \tab 344 \tab 160 \tab 77 \tab 30 \tab 19 \tab 22 \tab 3 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 562 \tab 332 \tab 528 \tab 329 \tab 144 \tab 61 \tab 35 \tab 18 \tab 32 \tab 3 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 536 \tab 274 \tab 569 \tab 301 \tab 167 \tab 56 \tab 33 \tab 12 \tab 23 \tab 3 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 704 \tab 419 \tab 658 \tab 394 \tab 198 \tab 88 \tab 40 \tab 14 \tab 15 \tab 8 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 797 \tab 459 \tab 733 \tab 467 \tab 213 \tab 92 \tab 51 \tab 25 \tab 22 \tab 8 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 663 \tab 349 \tab 623 \tab 383 \tab 166 \tab 74 \tab 47 \tab 23 \tab 16 \tab 4 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1163 \tab 646 \tab 1152 \tab 578 \tab 277 \tab 79 \tab 52 \tab 17 \tab 19 \tab - \tab 5 \tab 33 \tab 11 \tab 4032 \cr
   2022 \tab 1107 \tab 545 \tab 914 \tab 519 \tab 245 \tab 96 \tab 41 \tab 25 \tab 41 \tab - \tab 11 \tab - \tab - \tab 3544 \cr
   Total \tab 19927 \tab 11495 \tab 18138 \tab 11196 \tab 5753 \tab 2573 \tab 1328 \tab 718 \tab 1001 \tab 201 \tab 16 \tab 33 \tab 11 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{8} 8 or more
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
