% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_600.R
\name{china}
\alias{china}
\title{Liking for china}
\source{
General Social Survey https://gss.norc.org
}
\description{
china
}
\details{
Question 103. You will notice that the boxes on this card go from the highest position of "plus 5" for a country which you like very much, to the lowest position of "minus 5" for a country you dislike very much. How far up the scale or how far down the scale would you rate the following countries? f. China
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab not available in this year \tab +1 \tab +2 \tab +3 \tab +4 \tab +5 \tab -1 \tab -2 \tab -3 \tab -4 \tab -5 \tab dk, na, iap \tab iap \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 246 \tab 158 \tab 108 \tab 55 \tab 46 \tab 198 \tab 114 \tab 129 \tab 97 \tab 238 \tab 95 \tab - \tab 1484 \cr
   1975 \tab - \tab 226 \tab 120 \tab 105 \tab 43 \tab 38 \tab 228 \tab 123 \tab 134 \tab 115 \tab 237 \tab 121 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 251 \tab 105 \tab 95 \tab 55 \tab 45 \tab 235 \tab 121 \tab 141 \tab 92 \tab 252 \tab 138 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 333 \tab 217 \tab 214 \tab 102 \tab 84 \tab 246 \tab 138 \tab 124 \tab 105 \tab 129 \tab 168 \tab - \tab 1860 \cr
   1983 \tab - \tab 323 \tab 206 \tab 137 \tab 61 \tab 54 \tab 259 \tab 146 \tab 126 \tab 67 \tab 117 \tab 103 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 337 \tab 211 \tab 226 \tab 86 \tab 69 \tab 183 \tab 98 \tab 92 \tab 50 \tab 76 \tab 106 \tab - \tab 1534 \cr
   1986 \tab - \tab 352 \tab 223 \tab 217 \tab 77 \tab 67 \tab 157 \tab 90 \tab 86 \tab 52 \tab 83 \tab 66 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 249 \tab 125 \tab 135 \tab 68 \tab 64 \tab 114 \tab 50 \tab 24 \tab 25 \tab 57 \tab 570 \tab - \tab 1481 \cr
   1989 \tab - \tab 284 \tab 128 \tab 119 \tab 59 \tab 43 \tab 124 \tab 52 \tab 39 \tab 21 \tab 54 \tab 614 \tab - \tab 1537 \cr
   1990 \tab - \tab 199 \tab 102 \tab 86 \tab 34 \tab 41 \tab 137 \tab 67 \tab 51 \tab 37 \tab 109 \tab 509 \tab - \tab 1372 \cr
   1991 \tab - \tab 205 \tab 132 \tab 115 \tab 56 \tab 57 \tab 117 \tab 72 \tab 58 \tab 43 \tab 93 \tab 569 \tab - \tab 1517 \cr
   1993 \tab - \tab 247 \tab 115 \tab 101 \tab 42 \tab 53 \tab 138 \tab 96 \tab 76 \tab 45 \tab 103 \tab 590 \tab - \tab 1606 \cr
   1994 \tab - \tab 92 \tab 43 \tab 41 \tab 25 \tab 23 \tab 76 \tab 53 \tab 47 \tab 27 \tab 51 \tab 2514 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 47374 \tab 3344 \tab 1885 \tab 1699 \tab 763 \tab 684 \tab 2212 \tab 1220 \tab 1127 \tab 776 \tab 1599 \tab 6163 \tab 3544 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} +5
\item \code{1} +4
\item \code{2} +3
\item \code{3} +2
\item \code{4} +1
\item \code{5} -1
\item \code{6} -2
\item \code{7} -3
\item \code{8} -4
\item \code{9} -5
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
