% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_69.R
\name{chldmore}
\alias{chldmore}
\title{chldmore}
\source{
General Social Survey https://gss.norc.org
}
\description{
Expect more children
}
\details{
Question 212. Do you expect to have any (more) children?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab inappropriate \tab no \tab no answer \tab uncertain \tab yes \tab iap \tab Total \cr
   1972 \tab 334 \tab 882 \tab 28 \tab 87 \tab 282 \tab - \tab 1613 \cr
   1973 \tab - \tab - \tab - \tab - \tab - \tab 1504 \tab 1504 \cr
   1974 \tab 274 \tab 859 \tab 18 \tab 53 \tab 280 \tab - \tab 1484 \cr
   1975 \tab 289 \tab 821 \tab 21 \tab 68 \tab 291 \tab - \tab 1490 \cr
   1976 \tab 265 \tab 863 \tab 22 \tab 67 \tab 282 \tab - \tab 1499 \cr
   1977 \tab 132 \tab 998 \tab 39 \tab 62 \tab 299 \tab - \tab 1530 \cr
   1978 \tab 3 \tab 1112 \tab 5 \tab 61 \tab 351 \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab - \tab - \tab - \tab 1468 \tab 1468 \cr
   1982 \tab - \tab 1335 \tab 21 \tab 63 \tab 441 \tab - \tab 1860 \cr
   1983 \tab - \tab 1158 \tab 2 \tab 59 \tab 380 \tab - \tab 1599 \cr
   1984 \tab - \tab - \tab - \tab - \tab - \tab 1473 \tab 1473 \cr
   1985 \tab - \tab 1136 \tab 6 \tab 52 \tab 340 \tab - \tab 1534 \cr
   1986 \tab - \tab 1085 \tab 1 \tab 54 \tab 330 \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab - \tab - \tab - \tab 1819 \tab 1819 \cr
   1988 \tab - \tab 704 \tab 1 \tab 41 \tab 242 \tab 493 \tab 1481 \cr
   1989 \tab - \tab 740 \tab 4 \tab 36 \tab 226 \tab 531 \tab 1537 \cr
   1990 \tab - \tab 686 \tab - \tab 37 \tab 205 \tab 444 \tab 1372 \cr
   1991 \tab - \tab 753 \tab 4 \tab 46 \tab 221 \tab 493 \tab 1517 \cr
   1993 \tab - \tab 795 \tab 9 \tab 46 \tab 230 \tab 526 \tab 1606 \cr
   1994 \tab - \tab 375 \tab 3 \tab 25 \tab 104 \tab 2485 \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 1297 \tab 14302 \tab 184 \tab 857 \tab 4504 \tab 51246 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} uncertain
\item \code{4} inappropriate
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
