% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_69.R
\name{chldsoon}
\alias{chldsoon}
\title{Children expected in 5 years}
\source{
General Social Survey https://gss.norc.org
}
\description{
chldsoon
}
\details{
Question 212b. How many (more) in the next five years?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab 0 \tab 1 \tab 2 \tab 3 \tab 5 \tab iap \tab no answer \tab 4 \tab don't know \tab Total \cr
   1972 \tab 33 \tab 118 \tab 73 \tab 6 \tab 2 \tab 1303 \tab 78 \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab - \tab - \tab - \tab - \tab 1504 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 50 \tab 110 \tab 67 \tab 5 \tab 1 \tab 1186 \tab 38 \tab 3 \tab 24 \tab 1484 \cr
   1975 \tab 40 \tab 144 \tab 70 \tab 4 \tab - \tab 1178 \tab 36 \tab - \tab 18 \tab 1490 \cr
   1976 \tab 61 \tab 121 \tab 61 \tab 5 \tab - \tab 1195 \tab 34 \tab 1 \tab 21 \tab 1499 \cr
   1977 \tab 56 \tab 129 \tab 76 \tab 7 \tab 1 \tab 1192 \tab 57 \tab 2 \tab 10 \tab 1530 \cr
   1978 \tab 58 \tab 160 \tab 99 \tab 5 \tab 2 \tab 1176 \tab 22 \tab 1 \tab 9 \tab 1532 \cr
   1980 \tab - \tab - \tab - \tab - \tab - \tab 1468 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 108 \tab 179 \tab 101 \tab 15 \tab - \tab 1398 \tab 42 \tab 2 \tab 15 \tab 1860 \cr
   1983 \tab 53 \tab 195 \tab 104 \tab 14 \tab - \tab 1217 \tab 9 \tab 2 \tab 5 \tab 1599 \cr
   1984 \tab - \tab - \tab - \tab - \tab - \tab 1473 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab - \tab - \tab - \tab - \tab 1534 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab - \tab - \tab - \tab 1470 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab - \tab - \tab - \tab 1819 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab - \tab - \tab - \tab - \tab 1481 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab - \tab - \tab - \tab - \tab 1537 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab - \tab - \tab - \tab - \tab 1372 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 459 \tab 1156 \tab 651 \tab 61 \tab 6 \tab 69628 \tab 316 \tab 11 \tab 102 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
