% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{colath}
\alias{colath}
\title{Allow anti-religionist to teach}
\source{
General Social Survey https://gss.norc.org
}
\description{
colath
}
\details{
Question 76. There are always some people whose ideas are considered bad or dangerous by other people. For instance, somebody who is against churches and religion... b. Should such a person be allowed to teach in a college or university, or not?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab don't know \tab no answer \tab not allowed \tab yes, allowed to teach \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 72 \tab 3 \tab 894 \tab 644 \tab - \tab - \tab 1613 \cr
   1973 \tab 47 \tab 5 \tab 842 \tab 610 \tab - \tab - \tab 1504 \cr
   1974 \tab 53 \tab 2 \tab 810 \tab 619 \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab - \tab - \tab - \tab 1490 \tab - \tab 1490 \cr
   1976 \tab 33 \tab - \tab 847 \tab 619 \tab - \tab - \tab 1499 \cr
   1977 \tab 30 \tab 3 \tab 905 \tab 592 \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab - \tab - \tab - \tab 1532 \tab - \tab 1532 \cr
   1980 \tab 45 \tab 4 \tab 754 \tab 665 \tab - \tab - \tab 1468 \cr
   1982 \tab 74 \tab 3 \tab 978 \tab 805 \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab - \tab - \tab - \tab 1599 \tab - \tab 1599 \cr
   1984 \tab 48 \tab 4 \tab 748 \tab 673 \tab - \tab - \tab 1473 \cr
   1985 \tab 44 \tab 5 \tab 790 \tab 695 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab - \tab - \tab 1470 \tab - \tab 1470 \cr
   1987 \tab 66 \tab 8 \tab 915 \tab 830 \tab - \tab - \tab 1819 \cr
   1988 \tab 31 \tab 2 \tab 502 \tab 442 \tab 504 \tab - \tab 1481 \cr
   1989 \tab 47 \tab 1 \tab 458 \tab 527 \tab 504 \tab - \tab 1537 \cr
   1990 \tab 51 \tab 3 \tab 403 \tab 460 \tab 455 \tab - \tab 1372 \cr
   1991 \tab 35 \tab 3 \tab 440 \tab 515 \tab 524 \tab - \tab 1517 \cr
   1993 \tab 43 \tab 6 \tab 470 \tab 556 \tab 531 \tab - \tab 1606 \cr
   1994 \tab 84 \tab 7 \tab 871 \tab 1034 \tab 996 \tab - \tab 2992 \cr
   1996 \tab 91 \tab 3 \tab 769 \tab 1060 \tab 981 \tab - \tab 2904 \cr
   1998 \tab 99 \tab 7 \tab 707 \tab 1069 \tab 950 \tab - \tab 2832 \cr
   2000 \tab 88 \tab 7 \tab 725 \tab 1041 \tab 956 \tab - \tab 2817 \cr
   2002 \tab 17 \tab 1 \tab 350 \tab 556 \tab 1841 \tab - \tab 2765 \cr
   2004 \tab 24 \tab 1 \tab 305 \tab 568 \tab 1914 \tab - \tab 2812 \cr
   2006 \tab 53 \tab 2 \tab 731 \tab 1217 \tab 2507 \tab - \tab 4510 \cr
   2008 \tab 42 \tab 5 \tab 490 \tab 815 \tab 671 \tab - \tab 2023 \cr
   2010 \tab 36 \tab 5 \tab 473 \tab 767 \tab 763 \tab - \tab 2044 \cr
   2012 \tab 36 \tab 5 \tab 449 \tab 818 \tab 666 \tab - \tab 1974 \cr
   2014 \tab 31 \tab 8 \tab 556 \tab 1119 \tab 824 \tab - \tab 2538 \cr
   2016 \tab 41 \tab 9 \tab 610 \tab 1230 \tab 977 \tab - \tab 2867 \cr
   2018 \tab 45 \tab 3 \tab 505 \tab 1021 \tab 774 \tab - \tab 2348 \cr
   2021 \tab 9 \tab 2 \tab 805 \tab 1842 \tab 1357 \tab 17 \tab 4032 \cr
   2022 \tab 40 \tab 4 \tab 757 \tab 1532 \tab 1203 \tab 8 \tab 3544 \cr
   Total \tab 1455 \tab 121 \tab 19859 \tab 24941 \tab 25989 \tab 25 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{4} yes, allowed to teach
\item \code{5} not allowed
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
