% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_32.R
\name{colhomo}
\alias{colhomo}
\title{colhomo}
\source{
General Social Survey https://gss.norc.org
}
\description{
Allow homosexual to teach
}
\details{
Question 81. And what about a man who admits that he is a homosexual? b. Should such a person be allowed to teach in a college, or university, or not?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no answer \tab not allowed \tab yes, allowed to teach \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 64 \tab 2 \tab 727 \tab 711 \tab - \tab 1504 \cr
   1974 \tab - \tab 77 \tab 3 \tab 660 \tab 744 \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 48 \tab 3 \tab 668 \tab 780 \tab - \tab 1499 \cr
   1977 \tab - \tab 64 \tab 4 \tab 710 \tab 752 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 56 \tab 2 \tab 609 \tab 801 \tab - \tab 1468 \cr
   1982 \tab - \tab 94 \tab 9 \tab 755 \tab 1002 \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 59 \tab 5 \tab 545 \tab 864 \tab - \tab 1473 \cr
   1985 \tab - \tab 46 \tab 5 \tab 597 \tab 886 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 65 \tab 5 \tab 728 \tab 1021 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 45 \tab 5 \tab 375 \tab 552 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 48 \tab 3 \tab 328 \tab 654 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 43 \tab 3 \tab 297 \tab 574 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 36 \tab 3 \tab 326 \tab 628 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 47 \tab 5 \tab 290 \tab 733 \tab - \tab 1606 \cr
   1994 \tab 996 \tab 58 \tab 7 \tab 522 \tab 1409 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 57 \tab 2 \tab 424 \tab 1440 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 77 \tab 3 \tab 415 \tab 1387 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 76 \tab 6 \tab 368 \tab 1411 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 21 \tab 3 \tab 178 \tab 722 \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 9 \tab 3 \tab 176 \tab 710 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 37 \tab 9 \tab 411 \tab 1546 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 30 \tab 3 \tab 261 \tab 1058 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 20 \tab 6 \tab 191 \tab 1064 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 27 \tab 8 \tab 192 \tab 1081 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 20 \tab 8 \tab 194 \tab 1492 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 26 \tab 7 \tab 221 \tab 1636 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 22 \tab 10 \tab 190 \tab 1352 \tab - \tab 2348 \cr
   2021 \tab 1357 \tab 2 \tab 1 \tab 183 \tab 2473 \tab 16 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 29943 \tab 1274 \tab 133 \tab 11541 \tab 29483 \tab 16 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{4} yes, allowed to teach
\item \code{5} not allowed
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
