% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{colmil}
\alias{colmil}
\title{Allow militarist to teach}
\source{
General Social Survey https://gss.norc.org
}
\description{
colmil
}
\details{
Question 80. Consider a person who advocates doing away with elections and letting the military run the country. b. Should such a person be allowed to teach in a college or university, or not?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no answer \tab not allowed \tab yes, allowed to teach \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 55 \tab 5 \tab 886 \tab 553 \tab - \tab 1499 \cr
   1977 \tab - \tab 39 \tab 4 \tab 968 \tab 519 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 60 \tab 4 \tab 827 \tab 577 \tab - \tab 1468 \cr
   1982 \tab - \tab 106 \tab 12 \tab 1063 \tab 679 \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 44 \tab 6 \tab 819 \tab 604 \tab - \tab 1473 \cr
   1985 \tab - \tab 49 \tab 6 \tab 871 \tab 608 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 82 \tab 6 \tab 1038 \tab 693 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 34 \tab 6 \tab 575 \tab 362 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 52 \tab 4 \tab 562 \tab 415 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 38 \tab 4 \tab 481 \tab 394 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 51 \tab 5 \tab 513 \tab 424 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 47 \tab 4 \tab 514 \tab 510 \tab - \tab 1606 \cr
   1994 \tab 996 \tab 71 \tab 5 \tab 1010 \tab 910 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 94 \tab 4 \tab 900 \tab 925 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 92 \tab 4 \tab 820 \tab 966 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 84 \tab 8 \tab 879 \tab 890 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 22 \tab 1 \tab 391 \tab 510 \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 20 \tab 3 \tab 398 \tab 477 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 55 \tab 5 \tab 892 \tab 1051 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 50 \tab 3 \tab 617 \tab 682 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 31 \tab 5 \tab 519 \tab 726 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 36 \tab 3 \tab 529 \tab 740 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 39 \tab 11 \tab 670 \tab 994 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 53 \tab 8 \tab 725 \tab 1104 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 43 \tab 6 \tab 637 \tab 888 \tab - \tab 2348 \cr
   2021 \tab 1357 \tab 6 \tab - \tab 1394 \tab 1243 \tab 32 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 32931 \tab 1353 \tab 132 \tab 19498 \tab 18444 \tab 32 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{4} yes, allowed to teach
\item \code{5} not allowed
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
