% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{colrac}
\alias{colrac}
\title{colrac}
\source{
General Social Survey https://gss.norc.org
}
\description{
Allow racist to teach
}
\details{
Question 78. Or consider a person who believes that Blacks are genetically inferior. b. Should such a person be allowed to teach in a college or university, or not?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no answer \tab not allowed \tab yes, allowed to teach \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 53 \tab 4 \tab 836 \tab 606 \tab - \tab 1499 \cr
   1977 \tab - \tab 39 \tab 4 \tab 864 \tab 623 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 63 \tab 6 \tab 769 \tab 630 \tab - \tab 1468 \cr
   1982 \tab - \tab 84 \tab 4 \tab 1031 \tab 741 \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 50 \tab 5 \tab 820 \tab 598 \tab - \tab 1473 \cr
   1985 \tab - \tab 51 \tab 4 \tab 835 \tab 644 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 66 \tab 9 \tab 993 \tab 751 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 34 \tab 3 \tab 536 \tab 404 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 35 \tab 6 \tab 520 \tab 472 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 51 \tab 4 \tab 453 \tab 409 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 39 \tab 3 \tab 535 \tab 416 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 41 \tab 6 \tab 563 \tab 465 \tab - \tab 1606 \cr
   1994 \tab 996 \tab 72 \tab 7 \tab 1072 \tab 845 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 80 \tab 6 \tab 961 \tab 876 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 80 \tab 4 \tab 923 \tab 875 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 68 \tab 13 \tab 924 \tab 856 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 13 \tab 2 \tab 428 \tab 481 \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 18 \tab 2 \tab 458 \tab 420 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 47 \tab 5 \tab 1014 \tab 937 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 45 \tab 2 \tab 697 \tab 608 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 28 \tab 7 \tab 652 \tab 594 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 30 \tab 6 \tab 661 \tab 611 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 25 \tab 6 \tab 879 \tab 804 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 42 \tab 8 \tab 1000 \tab 840 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 43 \tab 8 \tab 872 \tab 651 \tab - \tab 2348 \cr
   2021 \tab 1357 \tab 10 \tab 1 \tab 1769 \tab 864 \tab 31 \tab 4032 \cr
   2022 \tab 1203 \tab 34 \tab 8 \tab 1455 \tab 838 \tab 6 \tab 3544 \cr
   Total \tab 30590 \tab 1241 \tab 143 \tab 22520 \tab 17859 \tab 37 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{4} yes, allowed to teach
\item \code{5} not allowed
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
