% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{commun}
\alias{commun}
\title{commun}
\source{
General Social Survey https://gss.norc.org
}
\description{
Feelings about communism
}
\details{
Question 101. Thinking about all the different kinds of governments in the world today, which of these statements comes closes to how you feel about Communism as a form of government? 1. It's the worst kind of all. 2. It's bad, but no worse than some others. 3. It's all right for some countries. 4. It's a good form of government.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab bad, not worst \tab don't know \tab good form \tab no answer \tab ok for some \tab worst kind \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 398 \tab 56 \tab 45 \tab 3 \tab 363 \tab 639 \tab 1504 \cr
   1974 \tab - \tab 389 \tab 42 \tab 45 \tab 4 \tab 278 \tab 726 \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 385 \tab 41 \tab 18 \tab 2 \tab 292 \tab 761 \tab 1499 \cr
   1977 \tab - \tab 373 \tab 30 \tab 21 \tab 4 \tab 300 \tab 802 \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 387 \tab 39 \tab 15 \tab 3 \tab 189 \tab 835 \tab 1468 \cr
   1982 \tab - \tab 453 \tab 82 \tab 34 \tab 7 \tab 261 \tab 1023 \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 373 \tab 35 \tab 29 \tab 6 \tab 155 \tab 875 \tab 1473 \cr
   1985 \tab - \tab 401 \tab 35 \tab 21 \tab 5 \tab 197 \tab 875 \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 512 \tab 66 \tab 23 \tab 12 \tab 266 \tab 940 \tab 1819 \cr
   1988 \tab 505 \tab 305 \tab 25 \tab 19 \tab 5 \tab 156 \tab 466 \tab 1481 \cr
   1989 \tab 504 \tab 354 \tab 25 \tab 12 \tab 2 \tab 147 \tab 493 \tab 1537 \cr
   1990 \tab 455 \tab 275 \tab 28 \tab 6 \tab 4 \tab 103 \tab 501 \tab 1372 \cr
   1991 \tab 524 \tab 328 \tab 28 \tab 8 \tab 3 \tab 134 \tab 492 \tab 1517 \cr
   1993 \tab 531 \tab 378 \tab 32 \tab 13 \tab 5 \tab 161 \tab 486 \tab 1606 \cr
   1994 \tab 2508 \tab 157 \tab 14 \tab 5 \tab - \tab 58 \tab 250 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 52741 \tab 5468 \tab 578 \tab 314 \tab 65 \tab 3060 \tab 10164 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} worst kind
\item \code{2} bad, not worst
\item \code{3} ok for some
\item \code{4} good form
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
