% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_78.R
\name{comprend}
\alias{comprend}
\title{comprend}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's understanding of questions
}
\details{
Question 247. Was respondent's understanding of the questions:
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab fair \tab good \tab no answer \tab poor \tab don't know \tab iap \tab Total \cr
   1972 \tab 312 \tab 1237 \tab 14 \tab 50 \tab - \tab - \tab 1613 \cr
   1973 \tab 280 \tab 1170 \tab 10 \tab 44 \tab - \tab - \tab 1504 \cr
   1974 \tab 278 \tab 1163 \tab 5 \tab 38 \tab - \tab - \tab 1484 \cr
   1975 \tab 251 \tab 1187 \tab 8 \tab 44 \tab - \tab - \tab 1490 \cr
   1976 \tab 280 \tab 1158 \tab 18 \tab 43 \tab - \tab - \tab 1499 \cr
   1977 \tab 247 \tab 1220 \tab 8 \tab 55 \tab - \tab - \tab 1530 \cr
   1978 \tab 299 \tab 1158 \tab 22 \tab 53 \tab - \tab - \tab 1532 \cr
   1980 \tab 269 \tab 1120 \tab 28 \tab 51 \tab - \tab - \tab 1468 \cr
   1982 \tab 360 \tab 1409 \tab 11 \tab 80 \tab - \tab - \tab 1860 \cr
   1983 \tab 255 \tab 1270 \tab 35 \tab 39 \tab - \tab - \tab 1599 \cr
   1984 \tab 230 \tab 1150 \tab 30 \tab 63 \tab - \tab - \tab 1473 \cr
   1985 \tab 204 \tab 1267 \tab 24 \tab 39 \tab - \tab - \tab 1534 \cr
   1986 \tab 234 \tab 1145 \tab 44 \tab 47 \tab - \tab - \tab 1470 \cr
   1987 \tab 311 \tab 1390 \tab 23 \tab 95 \tab - \tab - \tab 1819 \cr
   1988 \tab 258 \tab 1145 \tab 20 \tab 58 \tab - \tab - \tab 1481 \cr
   1989 \tab 249 \tab 1197 \tab 33 \tab 58 \tab - \tab - \tab 1537 \cr
   1990 \tab 210 \tab 1083 \tab 19 \tab 60 \tab - \tab - \tab 1372 \cr
   1991 \tab 207 \tab 1233 \tab 29 \tab 48 \tab - \tab - \tab 1517 \cr
   1993 \tab 240 \tab 1310 \tab 7 \tab 49 \tab - \tab - \tab 1606 \cr
   1994 \tab 423 \tab 2480 \tab 18 \tab 71 \tab - \tab - \tab 2992 \cr
   1996 \tab 391 \tab 2411 \tab 29 \tab 73 \tab - \tab - \tab 2904 \cr
   1998 \tab 360 \tab 2377 \tab 26 \tab 67 \tab 2 \tab - \tab 2832 \cr
   2000 \tab 391 \tab 2309 \tab 49 \tab 68 \tab - \tab - \tab 2817 \cr
   2002 \tab 303 \tab 2380 \tab - \tab 69 \tab 13 \tab - \tab 2765 \cr
   2004 \tab 312 \tab 2450 \tab 11 \tab 39 \tab - \tab - \tab 2812 \cr
   2006 \tab 516 \tab 3893 \tab 7 \tab 94 \tab - \tab - \tab 4510 \cr
   2008 \tab 286 \tab 1674 \tab 3 \tab 60 \tab - \tab - \tab 2023 \cr
   2010 \tab 341 \tab 1633 \tab 9 \tab 57 \tab 4 \tab - \tab 2044 \cr
   2012 \tab 298 \tab 1610 \tab 4 \tab 55 \tab 7 \tab - \tab 1974 \cr
   2014 \tab 462 \tab 2016 \tab 6 \tab 52 \tab 2 \tab - \tab 2538 \cr
   2016 \tab 417 \tab 2390 \tab 8 \tab 50 \tab 2 \tab - \tab 2867 \cr
   2018 \tab 331 \tab 1957 \tab 9 \tab 48 \tab 3 \tab - \tab 2348 \cr
   2021 \tab 66 \tab 254 \tab - \tab 5 \tab - \tab 3707 \tab 4032 \cr
   2022 \tab 373 \tab 1461 \tab - \tab 53 \tab 15 \tab 1642 \tab 3544 \cr
   Total \tab 10244 \tab 54307 \tab 567 \tab 1875 \tab 48 \tab 5349 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} good
\item \code{2} fair
\item \code{3} poor
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
