% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_262.R
\name{compuse}
\alias{compuse}
\title{compuse}
\source{
General Social Survey https://gss.norc.org
}
\description{
R use computer
}
\details{
Question 767. Do you personally ever use a computer at home, at work, or at some other location?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 454 \tab 993 \tab 10 \tab 1360 \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 913 \tab 6 \tab 1835 \tab 11 \tab - \tab 2765 \cr
   2004 \tab 1828 \tab 255 \tab 6 \tab 723 \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 614 \tab 288 \tab 1 \tab 1141 \tab - \tab - \tab 2044 \cr
   2012 \tab 672 \tab 253 \tab 3 \tab 1046 \tab - \tab - \tab 1974 \cr
   2014 \tab 863 \tab 289 \tab 2 \tab 1384 \tab - \tab - \tab 2538 \cr
   2016 \tab 979 \tab 374 \tab 3 \tab 1511 \tab - \tab - \tab 2867 \cr
   2018 \tab 789 \tab 312 \tab - \tab 1247 \tab - \tab - \tab 2348 \cr
   2021 \tab 1315 \tab 245 \tab 57 \tab 2406 \tab - \tab 9 \tab 4032 \cr
   2022 \tab 1168 \tab 430 \tab 11 \tab 1932 \tab 1 \tab 2 \tab 3544 \cr
   Total \tab 53331 \tab 4352 \tab 99 \tab 14585 \tab 12 \tab 11 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
