% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_52.R
\name{conarmy}
\alias{conarmy}
\title{Confidence in military}
\source{
General Social Survey https://gss.norc.org
}
\description{
conarmy
}
\details{
Question 165. I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? m. Military
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab don't know \tab hardly any \tab no answer \tab only some \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 475 \tab 41 \tab 241 \tab 6 \tab 741 \tab - \tab 1504 \cr
   1974 \tab - \tab 587 \tab 39 \tab 199 \tab 1 \tab 658 \tab - \tab 1484 \cr
   1975 \tab - \tab 524 \tab 69 \tab 213 \tab 3 \tab 681 \tab - \tab 1490 \cr
   1976 \tab - \tab 584 \tab 92 \tab 199 \tab 8 \tab 616 \tab - \tab 1499 \cr
   1977 \tab - \tab 554 \tab 48 \tab 157 \tab 4 \tab 767 \tab - \tab 1530 \cr
   1978 \tab - \tab 451 \tab 56 \tab 196 \tab 4 \tab 825 \tab - \tab 1532 \cr
   1980 \tab - \tab 405 \tab 61 \tab 237 \tab 1 \tab 764 \tab - \tab 1468 \cr
   1982 \tab - \tab 545 \tab 47 \tab 269 \tab 8 \tab 991 \tab - \tab 1860 \cr
   1983 \tab - \tab 468 \tab 48 \tab 207 \tab 3 \tab 873 \tab - \tab 1599 \cr
   1984 \tab 484 \tab 352 \tab 27 \tab 125 \tab 12 \tab 473 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 461 \tab 37 \tab 200 \tab 4 \tab 768 \tab - \tab 1470 \cr
   1987 \tab - \tab 610 \tab 60 \tab 232 \tab 5 \tab 912 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 338 \tab 32 \tab 134 \tab 2 \tab 491 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 335 \tab 42 \tab 139 \tab 3 \tab 516 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 293 \tab 26 \tab 121 \tab - \tab 459 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 605 \tab 18 \tab 60 \tab 3 \tab 331 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 441 \tab 18 \tab 116 \tab 5 \tab 477 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 737 \tab 46 \tab 246 \tab 11 \tab 971 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 716 \tab 62 \tab 213 \tab 2 \tab 932 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 689 \tab 53 \tab 237 \tab 5 \tab 927 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 727 \tab 68 \tab 184 \tab 10 \tab 907 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 492 \tab 14 \tab 63 \tab - \tab 343 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 495 \tab 4 \tab 74 \tab 1 \tab 302 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 929 \tab 31 \tab 256 \tab 1 \tab 772 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 681 \tab 17 \tab 153 \tab 2 \tab 512 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 720 \tab 19 \tab 129 \tab 2 \tab 507 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 726 \tab 20 \tab 109 \tab 1 \tab 482 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 843 \tab 14 \tab 157 \tab 3 \tab 670 \tab - \tab 2538 \cr
   2016 \tab 911 \tab 1027 \tab 16 \tab 147 \tab 3 \tab 763 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 935 \tab 17 \tab 103 \tab 4 \tab 504 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 1254 \tab 3 \tab 255 \tab - \tab 1147 \tab 13 \tab 4032 \cr
   2022 \tab 1173 \tab 1039 \tab 17 \tab 294 \tab 2 \tab 1017 \tab 2 \tab 3544 \cr
   Total \tab 23292 \tab 20038 \tab 1162 \tab 5665 \tab 119 \tab 22099 \tab 15 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
