% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_51.R
\name{conclerg}
\alias{conclerg}
\title{conclerg}
\source{
General Social Survey https://gss.norc.org
}
\description{
Confidence in organized religion
}
\details{
Question 165. I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? c. Organized religion.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab don't know \tab hardly any \tab no answer \tab only some \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 520 \tab 53 \tab 237 \tab 9 \tab 685 \tab - \tab 1504 \cr
   1974 \tab - \tab 656 \tab 31 \tab 160 \tab 3 \tab 634 \tab - \tab 1484 \cr
   1975 \tab - \tab 362 \tab 95 \tab 317 \tab 5 \tab 711 \tab - \tab 1490 \cr
   1976 \tab - \tab 457 \tab 94 \tab 273 \tab 8 \tab 667 \tab - \tab 1499 \cr
   1977 \tab - \tab 610 \tab 51 \tab 177 \tab 4 \tab 688 \tab - \tab 1530 \cr
   1978 \tab - \tab 468 \tab 58 \tab 278 \tab 6 \tab 722 \tab - \tab 1532 \cr
   1980 \tab - \tab 517 \tab 56 \tab 261 \tab 3 \tab 631 \tab - \tab 1468 \cr
   1982 \tab - \tab 602 \tab 60 \tab 281 \tab 7 \tab 910 \tab - \tab 1860 \cr
   1983 \tab - \tab 450 \tab 61 \tab 277 \tab 6 \tab 805 \tab - \tab 1599 \cr
   1984 \tab 484 \tab 304 \tab 34 \tab 184 \tab 10 \tab 457 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 369 \tab 50 \tab 308 \tab 3 \tab 740 \tab - \tab 1470 \cr
   1987 \tab - \tab 534 \tab 57 \tab 339 \tab 8 \tab 881 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 199 \tab 28 \tab 308 \tab 2 \tab 460 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 223 \tab 36 \tab 313 \tab 4 \tab 459 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 203 \tab 41 \tab 212 \tab 1 \tab 442 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 256 \tab 25 \tab 211 \tab 2 \tab 523 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 240 \tab 33 \tab 257 \tab 6 \tab 521 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 486 \tab 45 \tab 436 \tab 9 \tab 1035 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 482 \tab 91 \tab 365 \tab 3 \tab 984 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 512 \tab 61 \tab 358 \tab 6 \tab 974 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 519 \tab 81 \tab 356 \tab 6 \tab 934 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 168 \tab 14 \tab 217 \tab 2 \tab 511 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 206 \tab 17 \tab 208 \tab 3 \tab 442 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 487 \tab 54 \tab 429 \tab 5 \tab 1014 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 272 \tab 24 \tab 342 \tab 5 \tab 722 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 271 \tab 49 \tab 328 \tab 9 \tab 720 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 276 \tab 41 \tab 323 \tab 12 \tab 686 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 338 \tab 34 \tab 402 \tab 17 \tab 896 \tab - \tab 2538 \cr
   2016 \tab 911 \tab 379 \tab 30 \tab 503 \tab 12 \tab 1032 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 315 \tab 38 \tab 392 \tab 12 \tab 806 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 394 \tab 5 \tab 879 \tab 1 \tab 1377 \tab 16 \tab 4032 \cr
   2022 \tab 1173 \tab 336 \tab 43 \tab 813 \tab 10 \tab 1157 \tab 12 \tab 3544 \cr
   Total \tab 23292 \tab 12411 \tab 1490 \tab 10744 \tab 199 \tab 24226 \tab 28 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
