% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_51.R
\name{coneduc}
\alias{coneduc}
\title{coneduc}
\source{
General Social Survey https://gss.norc.org
}
\description{
Confidence in education
}
\details{
Question 165. I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? d. Education.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab don't know \tab hardly any \tab no answer \tab only some \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 553 \tab 21 \tab 123 \tab 9 \tab 798 \tab - \tab 1504 \cr
   1974 \tab - \tab 727 \tab 20 \tab 121 \tab 4 \tab 612 \tab - \tab 1484 \cr
   1975 \tab - \tab 460 \tab 26 \tab 190 \tab 2 \tab 812 \tab - \tab 1490 \cr
   1976 \tab - \tab 558 \tab 30 \tab 229 \tab 10 \tab 672 \tab - \tab 1499 \cr
   1977 \tab - \tab 620 \tab 14 \tab 135 \tab 4 \tab 757 \tab - \tab 1530 \cr
   1978 \tab - \tab 435 \tab 21 \tab 231 \tab 4 \tab 841 \tab - \tab 1532 \cr
   1980 \tab - \tab 438 \tab 30 \tab 183 \tab 2 \tab 815 \tab - \tab 1468 \cr
   1982 \tab - \tab 647 \tab 31 \tab 230 \tab 11 \tab 941 \tab - \tab 1860 \cr
   1983 \tab - \tab 457 \tab 28 \tab 209 \tab 5 \tab 900 \tab - \tab 1599 \cr
   1984 \tab 484 \tab 275 \tab 20 \tab 102 \tab 13 \tab 579 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 405 \tab 21 \tab 156 \tab 5 \tab 883 \tab - \tab 1470 \cr
   1987 \tab - \tab 654 \tab 25 \tab 153 \tab 4 \tab 983 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 293 \tab 18 \tab 86 \tab 3 \tab 597 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 313 \tab 16 \tab 108 \tab 4 \tab 594 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 241 \tab 15 \tab 110 \tab 4 \tab 529 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 306 \tab 17 \tab 136 \tab 1 \tab 557 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 234 \tab 17 \tab 192 \tab 5 \tab 609 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 499 \tab 23 \tab 349 \tab 6 \tab 1134 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 440 \tab 23 \tab 350 \tab 2 \tab 1110 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 511 \tab 23 \tab 315 \tab 7 \tab 1055 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 507 \tab 20 \tab 294 \tab 7 \tab 1068 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 227 \tab 6 \tab 142 \tab - \tab 537 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 240 \tab - \tab 127 \tab 2 \tab 507 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 550 \tab 11 \tab 318 \tab - \tab 1110 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 393 \tab 12 \tab 197 \tab - \tab 763 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 367 \tab 11 \tab 203 \tab 3 \tab 793 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 342 \tab 11 \tab 219 \tab - \tab 766 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 412 \tab 10 \tab 309 \tab - \tab 956 \tab - \tab 2538 \cr
   2016 \tab 911 \tab 493 \tab 7 \tab 357 \tab 1 \tab 1098 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 388 \tab 15 \tab 291 \tab 3 \tab 866 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 443 \tab 1 \tab 582 \tab - \tab 1633 \tab 13 \tab 4032 \cr
   2022 \tab 1173 \tab 428 \tab 12 \tab 567 \tab 1 \tab 1355 \tab 8 \tab 3544 \cr
   Total \tab 23292 \tab 13856 \tab 555 \tab 7314 \tab 122 \tab 27230 \tab 21 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
