% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_52.R
\name{conjudge}
\alias{conjudge}
\title{Confid. in united states supreme court}
\source{
General Social Survey https://gss.norc.org
}
\description{
conjudge
}
\details{
Question 165. I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? j. U.S. Supreme Court.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab don't know \tab hardly any \tab no answer \tab only some \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 471 \tab 50 \tab 231 \tab 7 \tab 745 \tab - \tab 1504 \cr
   1974 \tab - \tab 492 \tab 67 \tab 213 \tab 2 \tab 710 \tab - \tab 1484 \cr
   1975 \tab - \tab 457 \tab 64 \tab 276 \tab 5 \tab 688 \tab - \tab 1490 \cr
   1976 \tab - \tab 528 \tab 84 \tab 229 \tab 8 \tab 650 \tab - \tab 1499 \cr
   1977 \tab - \tab 543 \tab 62 \tab 165 \tab 8 \tab 752 \tab - \tab 1530 \cr
   1978 \tab - \tab 429 \tab 69 \tab 223 \tab 5 \tab 806 \tab - \tab 1532 \cr
   1980 \tab - \tab 361 \tab 87 \tab 286 \tab - \tab 734 \tab - \tab 1468 \cr
   1982 \tab - \tab 543 \tab 67 \tab 237 \tab 10 \tab 1003 \tab - \tab 1860 \cr
   1983 \tab - \tab 434 \tab 61 \tab 225 \tab 4 \tab 875 \tab - \tab 1599 \cr
   1984 \tab 484 \tab 325 \tab 35 \tab 121 \tab 11 \tab 497 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 434 \tab 55 \tab 206 \tab 10 \tab 765 \tab - \tab 1470 \cr
   1987 \tab - \tab 622 \tab 79 \tab 190 \tab 4 \tab 924 \tab - \tab 1819 \cr
   1988 \tab 485 \tab 344 \tab 44 \tab 105 \tab 4 \tab 499 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 356 \tab 49 \tab 111 \tab 2 \tab 517 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 315 \tab 39 \tab 114 \tab - \tab 431 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 373 \tab 45 \tab 124 \tab 5 \tab 470 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 322 \tab 45 \tab 141 \tab 3 \tab 546 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 604 \tab 69 \tab 328 \tab 7 \tab 1003 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 543 \tab 101 \tab 318 \tab 4 \tab 959 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 592 \tab 96 \tab 267 \tab 6 \tab 950 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 601 \tab 116 \tab 239 \tab 8 \tab 932 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 321 \tab 32 \tab 104 \tab 1 \tab 454 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 266 \tab 18 \tab 141 \tab - \tab 451 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 632 \tab 69 \tab 307 \tab 4 \tab 977 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 405 \tab 40 \tab 204 \tab - \tab 716 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 400 \tab 38 \tab 229 \tab 4 \tab 706 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 378 \tab 38 \tab 217 \tab 5 \tab 700 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 395 \tab 43 \tab 334 \tab 3 \tab 912 \tab - \tab 2538 \cr
   2016 \tab 911 \tab 488 \tab 37 \tab 389 \tab 4 \tab 1038 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 462 \tab 41 \tab 229 \tab 3 \tab 828 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 689 \tab 3 \tab 536 \tab 1 \tab 1437 \tab 6 \tab 4032 \cr
   2022 \tab 1173 \tab 373 \tab 29 \tab 894 \tab 9 \tab 1060 \tab 6 \tab 3544 \cr
   Total \tab 23293 \tab 14498 \tab 1772 \tab 7933 \tab 147 \tab 24735 \tab 12 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
