% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_51.R
\name{conlabor}
\alias{conlabor}
\title{Confidence in organized labor}
\source{
General Social Survey https://gss.norc.org
}
\description{
conlabor
}
\details{
Question 165. I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? f. Organized labor.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab don't know \tab hardly any \tab no answer \tab only some \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 232 \tab 62 \tab 384 \tab 9 \tab 817 \tab - \tab 1504 \cr
   1974 \tab - \tab 270 \tab 41 \tab 377 \tab 3 \tab 793 \tab - \tab 1484 \cr
   1975 \tab - \tab 150 \tab 95 \tab 436 \tab 2 \tab 807 \tab - \tab 1490 \cr
   1976 \tab - \tab 173 \tab 118 \tab 493 \tab 5 \tab 710 \tab - \tab 1499 \cr
   1977 \tab - \tab 225 \tab 59 \tab 483 \tab 6 \tab 757 \tab - \tab 1530 \cr
   1978 \tab - \tab 168 \tab 78 \tab 575 \tab 4 \tab 707 \tab - \tab 1532 \cr
   1980 \tab - \tab 220 \tab 85 \tab 434 \tab 2 \tab 727 \tab - \tab 1468 \cr
   1982 \tab - \tab 241 \tab 81 \tab 520 \tab 12 \tab 1006 \tab - \tab 1860 \cr
   1983 \tab - \tab 127 \tab 66 \tab 630 \tab 2 \tab 774 \tab - \tab 1599 \cr
   1984 \tab 484 \tab 84 \tab 24 \tab 353 \tab 11 \tab 517 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 117 \tab 79 \tab 576 \tab 5 \tab 693 \tab - \tab 1470 \cr
   1987 \tab - \tab 211 \tab 106 \tab 556 \tab 5 \tab 941 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 99 \tab 52 \tab 344 \tab 4 \tab 498 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 97 \tab 66 \tab 345 \tab 3 \tab 524 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 96 \tab 47 \tab 279 \tab 1 \tab 476 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 112 \tab 67 \tab 345 \tab 2 \tab 491 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 87 \tab 73 \tab 340 \tab 4 \tab 553 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 198 \tab 111 \tab 648 \tab 9 \tab 1045 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 215 \tab 162 \tab 567 \tab 2 \tab 979 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 210 \tab 155 \tab 556 \tab 4 \tab 986 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 239 \tab 185 \tab 500 \tab 10 \tab 962 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 101 \tab 50 \tab 210 \tab 1 \tab 550 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 103 \tab 44 \tab 249 \tab 2 \tab 478 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 223 \tab 111 \tab 568 \tab - \tab 1087 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 165 \tab 60 \tab 358 \tab 1 \tab 781 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 146 \tab 74 \tab 373 \tab 3 \tab 781 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 152 \tab 78 \tab 385 \tab 4 \tab 719 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 196 \tab 72 \tab 488 \tab 2 \tab 929 \tab - \tab 2538 \cr
   2016 \tab 911 \tab 249 \tab 78 \tab 440 \tab 4 \tab 1185 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 211 \tab 70 \tab 310 \tab 5 \tab 967 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 297 \tab 8 \tab 696 \tab - \tab 1655 \tab 16 \tab 4032 \cr
   2022 \tab 1173 \tab 311 \tab 84 \tab 582 \tab 4 \tab 1383 \tab 7 \tab 3544 \cr
   Total \tab 23292 \tab 5725 \tab 2541 \tab 14400 \tab 131 \tab 26278 \tab 23 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
