% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_51.R
\name{conpress}
\alias{conpress}
\title{conpress}
\source{
General Social Survey https://gss.norc.org
}
\description{
Confidence in press
}
\details{
Question 165. I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? g. Press.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab don't know \tab hardly any \tab no answer \tab only some \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 346 \tab 23 \tab 220 \tab 4 \tab 911 \tab - \tab 1504 \cr
   1974 \tab - \tab 383 \tab 18 \tab 259 \tab 3 \tab 821 \tab - \tab 1484 \cr
   1975 \tab - \tab 354 \tab 42 \tab 265 \tab 6 \tab 823 \tab - \tab 1490 \cr
   1976 \tab - \tab 424 \tab 27 \tab 263 \tab 9 \tab 776 \tab - \tab 1499 \cr
   1977 \tab - \tab 383 \tab 33 \tab 236 \tab 4 \tab 874 \tab - \tab 1530 \cr
   1978 \tab - \tab 307 \tab 27 \tab 301 \tab 4 \tab 893 \tab - \tab 1532 \cr
   1980 \tab - \tab 322 \tab 43 \tab 253 \tab 1 \tab 849 \tab - \tab 1468 \cr
   1982 \tab - \tab 330 \tab 45 \tab 364 \tab 10 \tab 1111 \tab - \tab 1860 \cr
   1983 \tab - \tab 215 \tab 31 \tab 375 \tab 4 \tab 974 \tab - \tab 1599 \cr
   1984 \tab 484 \tab 166 \tab 19 \tab 219 \tab 11 \tab 574 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 268 \tab 27 \tab 373 \tab 5 \tab 797 \tab - \tab 1470 \cr
   1987 \tab - \tab 342 \tab 47 \tab 411 \tab 3 \tab 1016 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 182 \tab 29 \tab 250 \tab 7 \tab 529 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 171 \tab 29 \tab 275 \tab 4 \tab 556 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 132 \tab 31 \tab 219 \tab 1 \tab 516 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 166 \tab 21 \tab 282 \tab 3 \tab 545 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 114 \tab 21 \tab 406 \tab 4 \tab 512 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 195 \tab 31 \tab 783 \tab 10 \tab 992 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 205 \tab 54 \tab 744 \tab 4 \tab 918 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 176 \tab 46 \tab 808 \tab 3 \tab 878 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 191 \tab 48 \tab 767 \tab 11 \tab 879 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 92 \tab 13 \tab 382 \tab - \tab 425 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 80 \tab 3 \tab 378 \tab - \tab 415 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 208 \tab 23 \tab 810 \tab 1 \tab 947 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 117 \tab 17 \tab 615 \tab - \tab 616 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 140 \tab 19 \tab 594 \tab 3 \tab 621 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 119 \tab 28 \tab 613 \tab - \tab 578 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 137 \tab 24 \tab 745 \tab - \tab 781 \tab - \tab 2538 \cr
   2016 \tab 911 \tab 159 \tab 17 \tab 963 \tab 2 \tab 815 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 199 \tab 13 \tab 713 \tab 2 \tab 636 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 306 \tab 2 \tab 1255 \tab - \tab 1093 \tab 16 \tab 4032 \cr
   2022 \tab 1173 \tab 203 \tab 21 \tab 1210 \tab 1 \tab 929 \tab 7 \tab 3544 \cr
   Total \tab 23292 \tab 7132 \tab 872 \tab 16351 \tab 120 \tab 24600 \tab 23 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
