% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_52.R
\name{consci}
\alias{consci}
\title{consci}
\source{
General Social Survey https://gss.norc.org
}
\description{
Confidence in scientific community
}
\details{
Question 165. I am going to name some institutions in this country. As far as the people running these institutions are concerned, would you say you have a great deal of confidence, only some confidence, or hardly any confidence at all in them? k. Scientific community.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab don't know \tab hardly any \tab no answer \tab only some \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 552 \tab 142 \tab 97 \tab 9 \tab 704 \tab - \tab 1504 \cr
   1974 \tab - \tab 667 \tab 157 \tab 99 \tab 3 \tab 558 \tab - \tab 1484 \cr
   1975 \tab - \tab 560 \tab 159 \tab 96 \tab 3 \tab 672 \tab - \tab 1490 \cr
   1976 \tab - \tab 638 \tab 172 \tab 112 \tab 13 \tab 564 \tab - \tab 1499 \cr
   1977 \tab - \tab 624 \tab 119 \tab 83 \tab 8 \tab 696 \tab - \tab 1530 \cr
   1978 \tab - \tab 553 \tab 126 \tab 111 \tab 5 \tab 737 \tab - \tab 1532 \cr
   1980 \tab - \tab 607 \tab 140 \tab 93 \tab 6 \tab 622 \tab - \tab 1468 \cr
   1982 \tab - \tab 645 \tab 179 \tab 133 \tab 17 \tab 886 \tab - \tab 1860 \cr
   1983 \tab - \tab 662 \tab 101 \tab 85 \tab 6 \tab 745 \tab - \tab 1599 \cr
   1984 \tab 484 \tab 436 \tab 56 \tab 54 \tab 13 \tab 430 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 569 \tab 84 \tab 113 \tab 7 \tab 697 \tab - \tab 1470 \cr
   1987 \tab - \tab 754 \tab 135 \tab 130 \tab 7 \tab 793 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 385 \tab 78 \tab 56 \tab 5 \tab 473 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 414 \tab 93 \tab 63 \tab 5 \tab 460 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 335 \tab 79 \tab 59 \tab 1 \tab 425 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 411 \tab 73 \tab 63 \tab 3 \tab 467 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 395 \tab 90 \tab 71 \tab 5 \tab 496 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 769 \tab 105 \tab 141 \tab 10 \tab 986 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 752 \tab 155 \tab 144 \tab 7 \tab 867 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 756 \tab 145 \tab 149 \tab 6 \tab 855 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 779 \tab 149 \tab 141 \tab 8 \tab 819 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 336 \tab 55 \tab 81 \tab - \tab 440 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 359 \tab 30 \tab 52 \tab 1 \tab 434 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 807 \tab 98 \tab 140 \tab 1 \tab 943 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 524 \tab 63 \tab 76 \tab 1 \tab 701 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 544 \tab 65 \tab 79 \tab 6 \tab 683 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 529 \tab 73 \tab 92 \tab 1 \tab 643 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 680 \tab 59 \tab 132 \tab 5 \tab 811 \tab - \tab 2538 \cr
   2016 \tab 911 \tab 791 \tab 69 \tab 117 \tab 3 \tab 976 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 681 \tab 55 \tab 100 \tab 2 \tab 725 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 1337 \tab 4 \tab 176 \tab - \tab 1141 \tab 14 \tab 4032 \cr
   2022 \tab 1173 \tab 930 \tab 51 \tab 296 \tab 3 \tab 1088 \tab 3 \tab 3544 \cr
   Total \tab 23292 \tab 19781 \tab 3159 \tab 3434 \tab 170 \tab 22537 \tab 17 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} only some
\item \code{3} hardly any
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
