% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{consent}
\alias{consent}
\title{consent}
\source{
General Social Survey https://gss.norc.org
}
\description{
Participation/recording consent
}
\details{
Question Your participation is voluntary. Your responses to this survey are considered a continuation of participation to the GSS survey you had completed earlier so the answers can be analyzed together, and all information that you provide will be kept in strict
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab no answer \tab r consents to participate in the survey, but does not want to be recorded \tab r consents to recording interview \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 2 \tab 324 \tab 1697 \tab 2023 \cr
   2010 \tab - \tab - \tab 357 \tab 1687 \tab 2044 \cr
   2012 \tab - \tab - \tab 244 \tab 1730 \tab 1974 \cr
   2014 \tab - \tab - \tab 463 \tab 2075 \tab 2538 \cr
   2016 \tab - \tab - \tab 267 \tab 2600 \tab 2867 \cr
   2018 \tab - \tab - \tab 166 \tab 2182 \tab 2348 \cr
   2021 \tab 2322 \tab - \tab 3 \tab 1707 \tab 4032 \cr
   2022 \tab 2695 \tab - \tab - \tab 849 \tab 3544 \cr
   Total \tab 56037 \tab 2 \tab 1824 \tab 14527 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} r consents to recording interview
\item \code{2} r consents to participate in the survey, but does not want to be recorded
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
