% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_78.R
\name{coop}
\alias{coop}
\title{coop}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's attitude toward interview
}
\details{
Question 246. In general, what was the respondent's attitude toward the interview?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab cooperative but not particularly interested \tab friendly and interested \tab hostile \tab impatient and restless \tab no answer \tab don't know \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 231 \tab 1220 \tab 6 \tab 33 \tab 14 \tab - \tab 1504 \cr
   1974 \tab - \tab 186 \tab 1249 \tab 6 \tab 39 \tab 4 \tab - \tab 1484 \cr
   1975 \tab - \tab 200 \tab 1229 \tab 9 \tab 41 \tab 11 \tab - \tab 1490 \cr
   1976 \tab - \tab 208 \tab 1237 \tab 11 \tab 30 \tab 13 \tab - \tab 1499 \cr
   1977 \tab - \tab 230 \tab 1235 \tab 5 \tab 51 \tab 9 \tab - \tab 1530 \cr
   1978 \tab - \tab 244 \tab 1221 \tab 2 \tab 41 \tab 24 \tab - \tab 1532 \cr
   1980 \tab - \tab 224 \tab 1160 \tab 2 \tab 47 \tab 35 \tab - \tab 1468 \cr
   1982 \tab - \tab 286 \tab 1486 \tab 7 \tab 69 \tab 12 \tab - \tab 1860 \cr
   1983 \tab - \tab 263 \tab 1265 \tab 6 \tab 54 \tab 11 \tab - \tab 1599 \cr
   1984 \tab - \tab 237 \tab 1164 \tab 4 \tab 60 \tab 8 \tab - \tab 1473 \cr
   1985 \tab - \tab 251 \tab 1210 \tab 6 \tab 52 \tab 15 \tab - \tab 1534 \cr
   1986 \tab - \tab 217 \tab 1188 \tab 6 \tab 43 \tab 16 \tab - \tab 1470 \cr
   1987 \tab - \tab 369 \tab 1341 \tab 13 \tab 77 \tab 19 \tab - \tab 1819 \cr
   1988 \tab - \tab 272 \tab 1136 \tab 10 \tab 52 \tab 11 \tab - \tab 1481 \cr
   1989 \tab - \tab 291 \tab 1151 \tab 12 \tab 61 \tab 22 \tab - \tab 1537 \cr
   1990 \tab - \tab 269 \tab 1024 \tab 5 \tab 64 \tab 10 \tab - \tab 1372 \cr
   1991 \tab - \tab 275 \tab 1139 \tab 13 \tab 71 \tab 19 \tab - \tab 1517 \cr
   1993 \tab - \tab 261 \tab 1271 \tab 5 \tab 65 \tab 4 \tab - \tab 1606 \cr
   1994 \tab - \tab 614 \tab 2200 \tab 13 \tab 151 \tab 14 \tab - \tab 2992 \cr
   1996 \tab - \tab 645 \tab 2073 \tab 11 \tab 148 \tab 27 \tab - \tab 2904 \cr
   1998 \tab - \tab 652 \tab 1940 \tab 25 \tab 179 \tab 34 \tab 2 \tab 2832 \cr
   2000 \tab - \tab 649 \tab 1978 \tab 12 \tab 126 \tab 52 \tab - \tab 2817 \cr
   2002 \tab - \tab 398 \tab 2232 \tab 10 \tab 111 \tab - \tab 14 \tab 2765 \cr
   2004 \tab - \tab 413 \tab 2311 \tab 4 \tab 72 \tab 12 \tab - \tab 2812 \cr
   2006 \tab - \tab 665 \tab 3686 \tab 9 \tab 143 \tab 7 \tab - \tab 4510 \cr
   2008 \tab - \tab 330 \tab 1613 \tab 4 \tab 73 \tab 3 \tab - \tab 2023 \cr
   2010 \tab - \tab 340 \tab 1600 \tab 6 \tab 86 \tab 8 \tab 4 \tab 2044 \cr
   2012 \tab - \tab 327 \tab 1560 \tab 4 \tab 74 \tab 4 \tab 5 \tab 1974 \cr
   2014 \tab - \tab 456 \tab 1982 \tab 1 \tab 84 \tab 6 \tab 9 \tab 2538 \cr
   2016 \tab - \tab 355 \tab 2417 \tab 3 \tab 75 \tab 8 \tab 9 \tab 2867 \cr
   2018 \tab - \tab 348 \tab 1916 \tab 4 \tab 70 \tab 9 \tab 1 \tab 2348 \cr
   2021 \tab 3707 \tab 20 \tab 300 \tab - \tab 5 \tab - \tab - \tab 4032 \cr
   2022 \tab 1642 \tab 272 \tab 1545 \tab 1 \tab 81 \tab 1 \tab 2 \tab 3544 \cr
   Total \tab 6962 \tab 10998 \tab 51279 \tab 235 \tab 2428 \tab 442 \tab 46 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} friendly and interested
\item \code{2} cooperative but not particularly interested
\item \code{3} impatient and restless
\item \code{4} hostile
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
