% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_554.R
\name{crack30}
\alias{crack30}
\title{crack30}
\source{
General Social Survey https://gss.norc.org
}
\description{
R last use crack cocaine
}
\details{
Question 1561. How long has it been since you last used "crack" cocaine in chunk or rock form?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab don't know \tab more than 12 months ago but within the past 3 years \tab more than 3 years ago \tab more than 30 days ago but within the past 12 months \tab no answer \tab within the past 30 days \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2629 \tab 2 \tab 18 \tab 54 \tab 15 \tab 93 \tab 6 \tab 2817 \cr
   2002 \tab 2638 \tab 2 \tab 14 \tab 77 \tab 6 \tab 21 \tab 7 \tab 2765 \cr
   2004 \tab 2690 \tab 1 \tab 15 \tab 78 \tab 7 \tab 15 \tab 6 \tab 2812 \cr
   2006 \tab 4327 \tab 1 \tab 22 \tab 108 \tab 17 \tab 19 \tab 16 \tab 4510 \cr
   2008 \tab 1899 \tab - \tab 10 \tab 92 \tab 4 \tab 12 \tab 6 \tab 2023 \cr
   2010 \tab 1913 \tab - \tab 20 \tab 82 \tab 7 \tab 16 \tab 6 \tab 2044 \cr
   2012 \tab 1845 \tab - \tab 10 \tab 86 \tab 4 \tab 27 \tab 2 \tab 1974 \cr
   2014 \tab 2403 \tab 1 \tab 19 \tab 75 \tab 6 \tab 31 \tab 3 \tab 2538 \cr
   2016 \tab 2755 \tab - \tab 9 \tab 85 \tab 3 \tab 11 \tab 4 \tab 2867 \cr
   2018 \tab 2250 \tab - \tab 4 \tab 68 \tab 9 \tab 15 \tab 2 \tab 2348 \cr
   2021 \tab 3901 \tab 3 \tab 5 \tab 90 \tab 11 \tab 21 \tab 1 \tab 4032 \cr
   2022 \tab 3416 \tab 3 \tab 8 \tab 78 \tab 3 \tab 33 \tab 3 \tab 3544 \cr
   Total \tab 70782 \tab 13 \tab 154 \tab 973 \tab 92 \tab 314 \tab 62 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} within the past 30 days
\item \code{2} more than 30 days ago but within the past 12 months
\item \code{3} more than 12 months ago but within the past 3 years
\item \code{4} more than 3 years ago
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
