% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_155.R
\name{datefrst}
\alias{datefrst}
\title{Month and yr began first job}
\source{
General Social Survey https://gss.norc.org
}
\description{
datefrst
}
\details{
Question 488. In what month and year did you begin this job? Report the month and year in which you actually began this job, even if you started the job before you completed your highest grade in school.
}
\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
