% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_10.R
\name{degree}
\alias{degree}
\title{R's highest degree}
\source{
General Social Survey https://gss.norc.org
}
\description{
degree
}
\details{
Question 19. If finished 9th-12th grade: Did you ever get a high school diploma or a GED certificate?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab associate/junior college \tab bachelor's \tab graduate \tab high school \tab less than high school \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 17 \tab 124 \tab 52 \tab 762 \tab 635 \tab 23 \tab - \tab - \tab 1613 \cr
   1973 \tab 20 \tab 132 \tab 65 \tab 720 \tab 552 \tab 15 \tab - \tab - \tab 1504 \cr
   1974 \tab 25 \tab 139 \tab 73 \tab 722 \tab 524 \tab 1 \tab - \tab - \tab 1484 \cr
   1975 \tab 33 \tab 140 \tab 50 \tab 732 \tab 534 \tab 1 \tab - \tab - \tab 1490 \cr
   1976 \tab 24 \tab 152 \tab 61 \tab 728 \tab 528 \tab 6 \tab - \tab - \tab 1499 \cr
   1977 \tab 34 \tab 139 \tab 74 \tab 741 \tab 536 \tab 4 \tab 2 \tab - \tab 1530 \cr
   1978 \tab 42 \tab 151 \tab 62 \tab 804 \tab 470 \tab 2 \tab 1 \tab - \tab 1532 \cr
   1980 \tab 45 \tab 158 \tab 71 \tab 745 \tab 445 \tab 3 \tab 1 \tab - \tab 1468 \cr
   1982 \tab 74 \tab 165 \tab 84 \tab 938 \tab 590 \tab 5 \tab 4 \tab - \tab 1860 \cr
   1983 \tab 61 \tab 202 \tab 80 \tab 835 \tab 419 \tab - \tab 2 \tab - \tab 1599 \cr
   1984 \tab 54 \tab 175 \tab 77 \tab 764 \tab 400 \tab 3 \tab - \tab - \tab 1473 \cr
   1985 \tab 60 \tab 175 \tab 90 \tab 795 \tab 414 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 48 \tab 192 \tab 80 \tab 750 \tab 399 \tab 1 \tab - \tab - \tab 1470 \cr
   1987 \tab 77 \tab 238 \tab 87 \tab 934 \tab 473 \tab 8 \tab 2 \tab - \tab 1819 \cr
   1988 \tab 65 \tab 179 \tab 89 \tab 785 \tab 362 \tab 1 \tab - \tab - \tab 1481 \cr
   1989 \tab 96 \tab 189 \tab 95 \tab 814 \tab 336 \tab 5 \tab 2 \tab - \tab 1537 \cr
   1990 \tab 75 \tab 197 \tab 87 \tab 726 \tab 282 \tab 5 \tab - \tab - \tab 1372 \cr
   1991 \tab 73 \tab 224 \tab 81 \tab 825 \tab 305 \tab 9 \tab - \tab - \tab 1517 \cr
   1993 \tab 97 \tab 253 \tab 118 \tab 840 \tab 294 \tab 2 \tab 2 \tab - \tab 1606 \cr
   1994 \tab 176 \tab 497 \tab 216 \tab 1586 \tab 507 \tab 7 \tab 3 \tab - \tab 2992 \cr
   1996 \tab 187 \tab 471 \tab 224 \tab 1567 \tab 448 \tab 5 \tab 2 \tab - \tab 2904 \cr
   1998 \tab 209 \tab 478 \tab 205 \tab 1500 \tab 430 \tab 10 \tab - \tab - \tab 2832 \cr
   2000 \tab 206 \tab 435 \tab 218 \tab 1501 \tab 439 \tab 16 \tab 2 \tab - \tab 2817 \cr
   2002 \tab 202 \tab 443 \tab 230 \tab 1485 \tab 400 \tab - \tab 5 \tab - \tab 2765 \cr
   2004 \tab 224 \tab 507 \tab 281 \tab 1435 \tab 364 \tab - \tab 1 \tab - \tab 2812 \cr
   2006 \tab 377 \tab 763 \tab 403 \tab 2273 \tab 691 \tab 3 \tab - \tab - \tab 4510 \cr
   2008 \tab 173 \tab 355 \tab 194 \tab 1003 \tab 297 \tab - \tab 1 \tab - \tab 2023 \cr
   2010 \tab 145 \tab 375 \tab 218 \tab 1001 \tab 305 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 151 \tab 354 \tab 205 \tab 976 \tab 288 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 186 \tab 472 \tab 281 \tab 1269 \tab 330 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 216 \tab 536 \tab 318 \tab 1461 \tab 328 \tab 8 \tab - \tab - \tab 2867 \cr
   2018 \tab 196 \tab 465 \tab 247 \tab 1178 \tab 262 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 370 \tab 1036 \tab 760 \tab 1597 \tab 246 \tab - \tab 14 \tab 9 \tab 4032 \cr
   2022 \tab 317 \tab 737 \tab 477 \tab 1654 \tab 359 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 4355 \tab 11248 \tab 5953 \tab 36446 \tab 14192 \tab 143 \tab 44 \tab 9 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} less than high school
\item \code{1} high school
\item \code{2} associate/junior college
\item \code{3} bachelor's
\item \code{4} graduate
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
