% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_34.R
\name{denom}
\alias{denom}
\title{denom}
\source{
General Social Survey https://gss.norc.org
}
\description{
Specific denomination
}
\details{
Question 104a. If Protestant: What specific denomination is that, if any?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllll}{
   year \tab baptist, don't know which \tab episcopal church \tab iap \tab lutheran, don't know which \tab methodist, don't know which \tab no answer \tab no denomination or non-denominational church \tab other, specify \tab presbyterian, don't know which \tab african methodist episcopal church \tab african methodist episcopal zion church \tab american baptist association \tab american baptist churches in the u.s.a. \tab american lutheran church \tab lutheran church - missouri synod \tab lutheran church in america \tab national baptist convention of america \tab national baptist convention u.s.a., inc. \tab other baptist churches \tab other lutheran churches \tab other methodist churches \tab other presbyterian churches \tab presbyterian church in the u.s. \tab southern baptist convention \tab united methodist church \tab united presbyterian church in the u.s.a. \tab wisconsin evangelical lutheran synod \tab don't know \tab evangelical lutheran \tab presbyterian, merged \tab skipped on web \tab Total \cr
   1972 \tab 324 \tab 33 \tab 577 \tab 139 \tab 232 \tab 7 \tab 36 \tab 185 \tab 80 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 312 \tab 41 \tab 564 \tab 124 \tab 198 \tab 2 \tab 43 \tab 163 \tab 57 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 322 \tab 38 \tab 529 \tab 113 \tab 190 \tab 2 \tab 42 \tab 174 \tab 74 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 309 \tab 44 \tab 513 \tab 139 \tab 175 \tab 3 \tab 49 \tab 182 \tab 76 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 303 \tab 49 \tab 546 \tab 104 \tab 162 \tab 2 \tab 60 \tab 195 \tab 78 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 324 \tab 37 \tab 519 \tab 136 \tab 193 \tab 7 \tab 58 \tab 190 \tab 66 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 318 \tab 39 \tab 549 \tab 115 \tab 190 \tab 5 \tab 46 \tab 208 \tab 62 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 307 \tab 41 \tab 529 \tab 92 \tab 170 \tab 3 \tab 52 \tab 195 \tab 79 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 478 \tab 52 \tab 590 \tab 126 \tab 210 \tab 13 \tab 72 \tab 244 \tab 75 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 303 \tab 37 \tab 625 \tab 143 \tab 147 \tab 6 \tab 45 \tab 229 \tab 64 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 95 \tab 28 \tab 529 \tab 14 \tab 18 \tab 12 \tab 57 \tab 220 \tab 23 \tab 6 \tab 2 \tab 20 \tab 16 \tab 35 \tab 31 \tab 5 \tab 7 \tab 7 \tab 43 \tab 7 \tab 12 \tab 5 \tab 16 \tab 107 \tab 126 \tab 31 \tab 1 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 67 \tab 42 \tab 574 \tab 17 \tab 25 \tab 6 \tab 48 \tab 215 \tab 15 \tab 11 \tab 1 \tab 35 \tab 8 \tab 39 \tab 31 \tab 1 \tab 5 \tab 4 \tab 61 \tab 8 \tab 7 \tab 8 \tab 13 \tab 144 \tab 125 \tab 21 \tab 3 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 73 \tab 38 \tab 546 \tab 13 \tab 19 \tab 4 \tab 53 \tab 224 \tab 17 \tab 5 \tab 1 \tab 18 \tab 11 \tab 42 \tab 36 \tab 12 \tab 10 \tab 2 \tab 55 \tab 8 \tab 13 \tab 4 \tab 16 \tab 126 \tab 107 \tab 13 \tab 4 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 173 \tab 36 \tab 558 \tab 17 \tab 20 \tab 10 \tab 61 \tab 299 \tab 20 \tab 19 \tab 6 \tab 29 \tab 23 \tab 27 \tab 22 \tab 15 \tab 25 \tab 18 \tab 44 \tab 2 \tab 12 \tab 8 \tab 15 \tab 205 \tab 115 \tab 37 \tab 3 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 72 \tab 31 \tab 574 \tab 13 \tab 15 \tab 4 \tab 65 \tab 216 \tab 19 \tab 5 \tab 1 \tab 29 \tab 10 \tab 15 \tab 19 \tab 12 \tab 3 \tab 5 \tab 42 \tab 6 \tab 9 \tab 7 \tab 12 \tab 153 \tab 109 \tab 25 \tab 3 \tab 1 \tab 6 \tab - \tab - \tab 1481 \cr
   1989 \tab 92 \tab 43 \tab 563 \tab 10 \tab 17 \tab 6 \tab 86 \tab 204 \tab 12 \tab 5 \tab 4 \tab 21 \tab 9 \tab 29 \tab 33 \tab 11 \tab 10 \tab 5 \tab 34 \tab 7 \tab 6 \tab 10 \tab 17 \tab 134 \tab 116 \tab 16 \tab 5 \tab - \tab 17 \tab 15 \tab - \tab 1537 \cr
   1990 \tab 59 \tab 34 \tab 505 \tab 21 \tab 16 \tab 7 \tab 78 \tab 211 \tab 13 \tab 11 \tab 6 \tab 21 \tab 4 \tab 32 \tab 23 \tab 8 \tab 6 \tab 2 \tab 17 \tab 4 \tab 3 \tab 2 \tab 13 \tab 122 \tab 93 \tab 25 \tab 7 \tab 1 \tab 12 \tab 16 \tab - \tab 1372 \cr
   1991 \tab 88 \tab 32 \tab 549 \tab 13 \tab 12 \tab 3 \tab 53 \tab 260 \tab 19 \tab 12 \tab 4 \tab 43 \tab 12 \tab 32 \tab 29 \tab 11 \tab 9 \tab 2 \tab 23 \tab 5 \tab 1 \tab 3 \tab 14 \tab 128 \tab 108 \tab 29 \tab 2 \tab - \tab 12 \tab 9 \tab - \tab 1517 \cr
   1993 \tab 85 \tab 34 \tab 572 \tab 8 \tab 23 \tab 11 \tab 84 \tab 235 \tab 8 \tab 8 \tab 4 \tab 21 \tab 7 \tab 20 \tab 43 \tab 14 \tab 8 \tab 3 \tab 31 \tab 11 \tab 5 \tab 16 \tab 15 \tab 157 \tab 130 \tab 23 \tab 7 \tab - \tab 14 \tab 9 \tab - \tab 1606 \cr
   1994 \tab 178 \tab 66 \tab 1207 \tab 22 \tab 32 \tab 18 \tab 125 \tab 338 \tab 16 \tab 8 \tab 4 \tab 44 \tab 16 \tab 48 \tab 60 \tab 9 \tab 9 \tab 9 \tab 76 \tab 20 \tab 21 \tab 19 \tab 36 \tab 284 \tab 215 \tab 51 \tab 12 \tab 1 \tab 28 \tab 20 \tab - \tab 2992 \cr
   1996 \tab 181 \tab 79 \tab 1236 \tab 25 \tab 23 \tab 10 \tab 99 \tab 304 \tab 11 \tab 11 \tab 8 \tab 57 \tab 28 \tab 51 \tab 48 \tab 15 \tab 12 \tab 8 \tab 69 \tab 15 \tab 12 \tab 19 \tab 34 \tab 273 \tab 190 \tab 29 \tab 11 \tab - \tab 33 \tab 13 \tab - \tab 2904 \cr
   1998 \tab 158 \tab 60 \tab 1245 \tab 12 \tab 30 \tab 46 \tab 139 \tab 266 \tab 7 \tab 16 \tab 5 \tab 43 \tab 16 \tab 34 \tab 47 \tab 15 \tab 9 \tab 15 \tab 83 \tab 23 \tab 6 \tab 10 \tab 30 \tab 244 \tab 192 \tab 23 \tab 15 \tab 1 \tab 25 \tab 17 \tab - \tab 2832 \cr
   2000 \tab 132 \tab 52 \tab 1254 \tab 17 \tab 30 \tab 16 \tab 161 \tab 287 \tab 8 \tab 19 \tab 7 \tab 42 \tab 25 \tab 36 \tab 37 \tab 13 \tab 14 \tab 13 \tab 75 \tab 12 \tab 14 \tab 18 \tab 31 \tab 256 \tab 182 \tab 21 \tab 5 \tab 1 \tab 25 \tab 14 \tab - \tab 2817 \cr
   2002 \tab 186 \tab 58 \tab 1225 \tab 52 \tab 34 \tab 17 \tab 203 \tab 338 \tab 42 \tab 11 \tab 2 \tab 34 \tab 18 \tab 29 \tab 34 \tab 15 \tab 10 \tab 4 \tab 22 \tab 2 \tab 1 \tab 3 \tab 16 \tab 208 \tab 158 \tab 15 \tab 11 \tab 8 \tab 6 \tab 3 \tab - \tab 2765 \cr
   2004 \tab 204 \tab 59 \tab 1239 \tab 49 \tab 31 \tab 16 \tab 263 \tab 348 \tab 27 \tab 10 \tab 5 \tab 41 \tab 11 \tab 12 \tab 30 \tab 7 \tab 8 \tab 6 \tab 18 \tab 4 \tab 1 \tab 3 \tab 18 \tab 210 \tab 144 \tab 14 \tab 9 \tab 4 \tab 16 \tab 5 \tab - \tab 2812 \cr
   2006 \tab 322 \tab 78 \tab 2054 \tab 44 \tab 49 \tab 28 \tab 370 \tab 581 \tab 60 \tab 11 \tab 10 \tab 45 \tab 25 \tab 34 \tab 45 \tab 13 \tab 18 \tab 10 \tab 38 \tab 3 \tab 6 \tab 10 \tab 16 \tab 324 \tab 219 \tab 21 \tab 20 \tab 17 \tab 22 \tab 17 \tab - \tab 4510 \cr
   2008 \tab 150 \tab 43 \tab 907 \tab 20 \tab 27 \tab 9 \tab 191 \tab 289 \tab 31 \tab 8 \tab 1 \tab 20 \tab 12 \tab 11 \tab 12 \tab 3 \tab 9 \tab 1 \tab 5 \tab 1 \tab 2 \tab 3 \tab 6 \tab 132 \tab 93 \tab 15 \tab 8 \tab 7 \tab 6 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 184 \tab 36 \tab 969 \tab 35 \tab 30 \tab 13 \tab 198 \tab 209 \tab 29 \tab 3 \tab - \tab 13 \tab 5 \tab 4 \tab 13 \tab 7 \tab 6 \tab 2 \tab 12 \tab 3 \tab 3 \tab 5 \tab 4 \tab 133 \tab 78 \tab 12 \tab 6 \tab 8 \tab 14 \tab 10 \tab - \tab 2044 \cr
   2012 \tab 129 \tab 32 \tab 935 \tab 21 \tab 21 \tab 3 \tab 222 \tab 230 \tab 17 \tab 7 \tab 3 \tab 18 \tab 18 \tab 10 \tab 16 \tab 8 \tab 2 \tab 1 \tab 23 \tab 2 \tab 2 \tab 2 \tab 5 \tab 126 \tab 83 \tab 3 \tab 8 \tab 4 \tab 17 \tab 6 \tab - \tab 1974 \cr
   2014 \tab 151 \tab 39 \tab 1264 \tab 29 \tab 17 \tab 15 \tab 293 \tab 260 \tab 30 \tab 8 \tab 4 \tab 24 \tab 16 \tab 10 \tab 25 \tab 5 \tab 9 \tab 3 \tab 20 \tab 3 \tab 4 \tab 3 \tab 8 \tab 145 \tab 110 \tab 9 \tab 4 \tab 3 \tab 16 \tab 11 \tab - \tab 2538 \cr
   2016 \tab 189 \tab 40 \tab 1432 \tab 25 \tab 21 \tab 16 \tab 336 \tab 301 \tab 25 \tab 6 \tab 4 \tab 33 \tab 26 \tab 10 \tab 33 \tab 11 \tab 7 \tab 5 \tab 25 \tab 6 \tab 4 \tab 6 \tab 3 \tab 156 \tab 94 \tab 9 \tab 6 \tab 8 \tab 20 \tab 10 \tab - \tab 2867 \cr
   2018 \tab 147 \tab 26 \tab 1163 \tab 30 \tab 26 \tab 20 \tab 266 \tab 248 \tab 30 \tab 6 \tab 4 \tab 24 \tab 16 \tab 11 \tab 20 \tab 1 \tab 4 \tab 3 \tab 20 \tab 3 \tab 5 \tab 6 \tab 9 \tab 122 \tab 94 \tab 1 \tab 7 \tab 13 \tab 13 \tab 10 \tab - \tab 2348 \cr
   2021 \tab 205 \tab 55 \tab 2242 \tab 56 \tab 26 \tab 46 \tab 551 \tab 220 \tab 54 \tab 15 \tab 2 \tab 21 \tab 24 \tab 22 \tab 34 \tab 5 \tab 5 \tab 6 \tab 26 \tab 8 \tab - \tab 8 \tab 17 \tab 122 \tab 148 \tab 5 \tab 6 \tab 15 \tab 30 \tab 19 \tab 39 \tab 4032 \cr
   2022 \tab 175 \tab 44 \tab 2010 \tab 43 \tab 22 \tab 32 \tab 498 \tab 206 \tab 24 \tab 9 \tab 3 \tab 37 \tab 22 \tab 9 \tab 32 \tab 8 \tab 6 \tab 4 \tab 25 \tab 1 \tab 5 \tab 5 \tab 9 \tab 140 \tab 130 \tab 10 \tab 1 \tab - \tab 14 \tab 11 \tab 9 \tab 3544 \cr
   Total \tab 6795 \tab 1496 \tab 30893 \tab 1837 \tab 2451 \tab 418 \tab 5003 \tab 8474 \tab 1268 \tab 230 \tab 91 \tab 733 \tab 378 \tab 602 \tab 753 \tab 224 \tab 211 \tab 138 \tab 887 \tab 164 \tab 154 \tab 183 \tab 373 \tab 4151 \tab 3159 \tab 458 \tab 164 \tab 92 \tab 346 \tab 216 \tab 48 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{10} american baptist association
\item \code{11} american baptist churches in the u.s.a.
\item \code{12} national baptist convention of america
\item \code{13} national baptist convention u.s.a., inc.
\item \code{14} southern baptist convention
\item \code{15} other baptist churches
\item \code{18} baptist, don't know which
\item \code{20} african methodist episcopal church
\item \code{21} african methodist episcopal zion church
\item \code{22} united methodist church
\item \code{23} other methodist churches
\item \code{28} methodist, don't know which
\item \code{30} american lutheran church
\item \code{31} lutheran church in america
\item \code{32} lutheran church - missouri synod
\item \code{33} wisconsin evangelical lutheran synod
\item \code{34} other lutheran churches
\item \code{35} evangelical lutheran
\item \code{38} lutheran, don't know which
\item \code{40} presbyterian church in the u.s.
\item \code{41} united presbyterian church in the u.s.a.
\item \code{42} other presbyterian churches
\item \code{43} presbyterian, merged
\item \code{48} presbyterian, don't know which
\item \code{50} episcopal church
\item \code{60} other, specify
\item \code{70} no denomination or non-denominational church
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
