% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_40.R
\name{denom16}
\alias{denom16}
\title{denom16}
\source{
General Social Survey https://gss.norc.org
}
\description{
Denomination in which r was raised
}
\details{
Question 115a. If Protestant: What specific denomination is that, if any?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllll}{
   year \tab iap \tab baptist, don't know which \tab episcopal church \tab lutheran, don't know which \tab methodist, don't know which \tab no answer \tab no denomination or non-denominational church \tab other, specify \tab presbyterian, don't know which \tab don't know \tab african methodist episcopal church \tab american baptist association \tab american baptist churches in the u.s.a. \tab american lutheran church \tab lutheran church - missouri synod \tab lutheran church in america \tab national baptist convention of america \tab national baptist convention u.s.a., inc. \tab other baptist churches \tab other lutheran churches \tab other methodist churches \tab other presbyterian churches \tab presbyterian church in the u.s. \tab southern baptist convention \tab united methodist church \tab united presbyterian church in the u.s.a. \tab african methodist episcopal zion church \tab wisconsin evangelical lutheran synod \tab evangelical lutheran \tab presbyterian, merged \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 523 \tab 354 \tab 41 \tab 131 \tab 198 \tab 7 \tab 26 \tab 151 \tab 73 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 500 \tab 366 \tab 31 \tab 119 \tab 217 \tab 2 \tab 22 \tab 146 \tab 81 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 458 \tab 343 \tab 39 \tab 149 \tab 235 \tab 1 \tab 30 \tab 167 \tab 67 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 526 \tab 340 \tab 38 \tab 107 \tab 196 \tab 3 \tab 27 \tab 167 \tab 95 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 482 \tab 367 \tab 36 \tab 145 \tab 221 \tab 7 \tab 31 \tab 163 \tab 78 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 508 \tab 371 \tab 29 \tab 127 \tab 222 \tab 9 \tab 32 \tab 171 \tab 63 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 483 \tab 330 \tab 37 \tab 105 \tab 224 \tab 5 \tab 25 \tab 168 \tab 91 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 556 \tab 550 \tab 49 \tab 128 \tab 272 \tab 10 \tab 33 \tab 181 \tab 81 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 594 \tab 340 \tab 40 \tab 143 \tab 183 \tab 10 \tab 26 \tab 201 \tab 62 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 499 \tab 108 \tab 26 \tab 22 \tab 33 \tab 28 \tab 36 \tab 189 \tab 31 \tab - \tab 15 \tab 19 \tab 16 \tab 34 \tab 27 \tab 3 \tab 8 \tab 6 \tab 40 \tab 6 \tab 22 \tab 8 \tab 17 \tab 130 \tab 127 \tab 23 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 551 \tab 79 \tab 39 \tab 27 \tab 37 \tab 5 \tab 23 \tab 199 \tab 27 \tab - \tab 15 \tab 33 \tab 10 \tab 35 \tab 25 \tab 1 \tab 4 \tab 4 \tab 62 \tab 5 \tab 18 \tab 8 \tab 11 \tab 156 \tab 129 \tab 26 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 515 \tab 105 \tab 35 \tab 25 \tab 44 \tab 5 \tab 26 \tab 193 \tab 22 \tab 2 \tab 6 \tab 20 \tab 12 \tab 35 \tab 32 \tab 11 \tab 12 \tab 3 \tab 61 \tab 11 \tab 16 \tab 5 \tab 10 \tab 133 \tab 116 \tab 10 \tab - \tab 5 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 500 \tab 224 \tab 31 \tab 20 \tab 48 \tab 34 \tab 26 \tab 252 \tab 26 \tab 1 \tab 26 \tab 26 \tab 26 \tab 26 \tab 23 \tab 7 \tab 23 \tab 18 \tab 43 \tab 4 \tab 17 \tab 5 \tab 18 \tab 228 \tab 127 \tab 33 \tab 3 \tab 4 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 529 \tab 96 \tab 31 \tab 27 \tab 34 \tab 3 \tab 30 \tab 171 \tab 36 \tab - \tab 6 \tab 31 \tab 9 \tab 20 \tab 18 \tab 14 \tab 5 \tab 6 \tab 38 \tab 7 \tab 17 \tab 7 \tab 14 \tab 178 \tab 126 \tab 25 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 567 \tab 107 \tab 32 \tab 13 \tab 39 \tab 6 \tab 61 \tab 186 \tab 18 \tab - \tab 5 \tab 26 \tab 6 \tab 37 \tab 31 \tab 12 \tab 11 \tab 5 \tab 29 \tab 11 \tab 7 \tab 8 \tab 22 \tab 137 \tab 112 \tab 18 \tab 6 \tab 8 \tab 9 \tab 8 \tab - \tab 1537 \cr
   1990 \tab 487 \tab 85 \tab 30 \tab 27 \tab 36 \tab 11 \tab 40 \tab 190 \tab 18 \tab - \tab 9 \tab 26 \tab 7 \tab 32 \tab 25 \tab 9 \tab 7 \tab 2 \tab 16 \tab 4 \tab 6 \tab 3 \tab 15 \tab 129 \tab 108 \tab 22 \tab 5 \tab 5 \tab 9 \tab 9 \tab - \tab 1372 \cr
   1991 \tab 548 \tab 111 \tab 31 \tab 21 \tab 31 \tab 6 \tab 37 \tab 193 \tab 21 \tab - \tab 15 \tab 47 \tab 14 \tab 30 \tab 32 \tab 8 \tab 9 \tab 1 \tab 14 \tab 7 \tab 5 \tab 4 \tab 12 \tab 147 \tab 125 \tab 25 \tab 3 \tab 5 \tab 6 \tab 9 \tab - \tab 1517 \cr
   1993 \tab 557 \tab 93 \tab 41 \tab 15 \tab 44 \tab 15 \tab 37 \tab 237 \tab 15 \tab - \tab 7 \tab 21 \tab 12 \tab 26 \tab 40 \tab 14 \tab 7 \tab 4 \tab 24 \tab 11 \tab 11 \tab 16 \tab 18 \tab 171 \tab 115 \tab 26 \tab 8 \tab 5 \tab 5 \tab 11 \tab - \tab 1606 \cr
   1994 \tab 1139 \tab 208 \tab 68 \tab 36 \tab 57 \tab 21 \tab 69 \tab 309 \tab 28 \tab 2 \tab 16 \tab 61 \tab 21 \tab 56 \tab 60 \tab 13 \tab 11 \tab 15 \tab 56 \tab 24 \tab 26 \tab 14 \tab 42 \tab 334 \tab 213 \tab 42 \tab 7 \tab 10 \tab 21 \tab 13 \tab - \tab 2992 \cr
   1996 \tab 1165 \tab 215 \tab 73 \tab 28 \tab 65 \tab 16 \tab 60 \tab 271 \tab 23 \tab 2 \tab 19 \tab 64 \tab 27 \tab 61 \tab 45 \tab 15 \tab 13 \tab 7 \tab 52 \tab 14 \tab 16 \tab 21 \tab 35 \tab 288 \tab 227 \tab 27 \tab 11 \tab 11 \tab 19 \tab 14 \tab - \tab 2904 \cr
   1998 \tab 1147 \tab 220 \tab 68 \tab 13 \tab 65 \tab 48 \tab 60 \tab 251 \tab 27 \tab 3 \tab 19 \tab 42 \tab 14 \tab 44 \tab 51 \tab 13 \tab 7 \tab 17 \tab 64 \tab 28 \tab 11 \tab 22 \tab 31 \tab 271 \tab 210 \tab 35 \tab 12 \tab 14 \tab 16 \tab 9 \tab - \tab 2832 \cr
   2000 \tab 1172 \tab 190 \tab 47 \tab 21 \tab 54 \tab 25 \tab 91 \tab 272 \tab 15 \tab 3 \tab 20 \tab 44 \tab 21 \tab 35 \tab 46 \tab 16 \tab 10 \tab 16 \tab 59 \tab 21 \tab 11 \tab 20 \tab 29 \tab 305 \tab 202 \tab 26 \tab 11 \tab 8 \tab 12 \tab 15 \tab - \tab 2817 \cr
   2002 \tab 1159 \tab 229 \tab 70 \tab 68 \tab 62 \tab 20 \tab 107 \tab 317 \tab 56 \tab 6 \tab 9 \tab 40 \tab 16 \tab 31 \tab 36 \tab 15 \tab 10 \tab 5 \tab 24 \tab 7 \tab 1 \tab 4 \tab 21 \tab 243 \tab 167 \tab 18 \tab 1 \tab 9 \tab 9 \tab 5 \tab - \tab 2765 \cr
   2004 \tab 1196 \tab 234 \tab 57 \tab 59 \tab 51 \tab 10 \tab 152 \tab 329 \tab 47 \tab 5 \tab 17 \tab 62 \tab 14 \tab 18 \tab 29 \tab 7 \tab 12 \tab 8 \tab 19 \tab 9 \tab 3 \tab 2 \tab 16 \tab 225 \tab 179 \tab 20 \tab 6 \tab 11 \tab 11 \tab 4 \tab - \tab 2812 \cr
   2006 \tab 2863 \tab 254 \tab 61 \tab 59 \tab 70 \tab 2 \tab 100 \tab 356 \tab 66 \tab 26 \tab 15 \tab 45 \tab 16 \tab 24 \tab 34 \tab 6 \tab 11 \tab 5 \tab 18 \tab 5 \tab 6 \tab 3 \tab 10 \tab 256 \tab 144 \tab 17 \tab 4 \tab 14 \tab 12 \tab 8 \tab - \tab 4510 \cr
   2008 \tab 891 \tab 205 \tab 40 \tab 29 \tab 48 \tab 12 \tab 86 \tab 269 \tab 45 \tab 9 \tab 8 \tab 26 \tab 11 \tab 13 \tab 21 \tab 4 \tab 10 \tab 2 \tab 6 \tab 1 \tab - \tab - \tab 11 \tab 156 \tab 95 \tab 13 \tab 4 \tab 6 \tab 2 \tab - \tab - \tab 2023 \cr
   2010 \tab 888 \tab 220 \tab 48 \tab 43 \tab 42 \tab 10 \tab 143 \tab 182 \tab 54 \tab 13 \tab 6 \tab 16 \tab 4 \tab 8 \tab 15 \tab 12 \tab 6 \tab 4 \tab 27 \tab 4 \tab 5 \tab 7 \tab 6 \tab 157 \tab 91 \tab 8 \tab 1 \tab 5 \tab 8 \tab 11 \tab - \tab 2044 \cr
   2012 \tab 905 \tab 175 \tab 38 \tab 39 \tab 40 \tab 4 \tab 122 \tab 198 \tab 30 \tab 5 \tab 10 \tab 31 \tab 19 \tab 10 \tab 21 \tab 10 \tab 3 \tab 1 \tab 16 \tab 2 \tab 2 \tab 3 \tab 5 \tab 154 \tab 102 \tab 4 \tab 2 \tab 6 \tab 11 \tab 6 \tab - \tab 1974 \cr
   2014 \tab 1185 \tab 213 \tab 44 \tab 49 \tab 38 \tab 12 \tab 156 \tab 293 \tab 45 \tab 11 \tab 10 \tab 30 \tab 23 \tab 13 \tab 29 \tab 2 \tab 9 \tab 3 \tab 20 \tab 3 \tab 7 \tab 4 \tab 9 \tab 177 \tab 110 \tab 11 \tab 6 \tab 6 \tab 11 \tab 9 \tab - \tab 2538 \cr
   2016 \tab 1316 \tab 256 \tab 54 \tab 45 \tab 50 \tab 11 \tab 191 \tab 302 \tab 50 \tab 23 \tab 10 \tab 40 \tab 26 \tab 18 \tab 34 \tab 6 \tab 12 \tab 9 \tab 29 \tab 9 \tab 15 \tab 4 \tab 9 \tab 189 \tab 122 \tab 11 \tab 3 \tab 7 \tab 15 \tab 1 \tab - \tab 2867 \cr
   2018 \tab 1090 \tab 222 \tab 37 \tab 54 \tab 39 \tab 12 \tab 151 \tab 237 \tab 33 \tab 20 \tab 10 \tab 28 \tab 19 \tab 12 \tab 21 \tab 2 \tab 5 \tab 1 \tab 34 \tab 4 \tab 5 \tab 6 \tab 13 \tab 168 \tab 100 \tab 5 \tab 1 \tab 5 \tab 9 \tab 5 \tab - \tab 2348 \cr
   2021 \tab 2108 \tab 287 \tab 74 \tab 97 \tab 51 \tab 4 \tab 289 \tab 291 \tab 101 \tab 29 \tab 24 \tab 32 \tab 28 \tab 17 \tab 45 \tab 10 \tab 6 \tab 10 \tab 34 \tab 6 \tab 9 \tab 6 \tab 18 \tab 169 \tab 222 \tab 11 \tab 3 \tab 7 \tab 26 \tab 11 \tab 7 \tab 4032 \cr
   2022 \tab 1749 \tab 269 \tab 63 \tab 63 \tab 40 \tab 18 \tab 324 \tab 281 \tab 54 \tab 78 \tab 10 \tab 51 \tab 25 \tab 18 \tab 31 \tab 7 \tab 10 \tab 5 \tab 17 \tab 3 \tab 5 \tab 4 \tab 19 \tab 192 \tab 142 \tab 12 \tab 4 \tab 6 \tab 17 \tab 6 \tab 21 \tab 3544 \cr
   Total \tab 30969 \tab 7766 \tab 1478 \tab 2054 \tab 3086 \tab 392 \tab 2669 \tab 7483 \tab 1579 \tab 239 \tab 307 \tab 861 \tab 396 \tab 653 \tab 771 \tab 217 \tab 221 \tab 157 \tab 802 \tab 206 \tab 241 \tab 184 \tab 411 \tab 4693 \tab 3411 \tab 468 \tab 104 \tab 162 \tab 228 \tab 154 \tab 28 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{10} american baptist association
\item \code{11} american baptist churches in the u.s.a.
\item \code{12} national baptist convention of america
\item \code{13} national baptist convention u.s.a., inc.
\item \code{14} southern baptist convention
\item \code{15} other baptist churches
\item \code{18} baptist, don't know which
\item \code{20} african methodist episcopal church
\item \code{21} african methodist episcopal zion church
\item \code{22} united methodist church
\item \code{23} other methodist churches
\item \code{28} methodist, don't know which
\item \code{30} american lutheran church
\item \code{31} lutheran church in america
\item \code{32} lutheran church - missouri synod
\item \code{33} wisconsin evangelical lutheran synod
\item \code{34} other lutheran churches
\item \code{35} evangelical lutheran
\item \code{38} lutheran, don't know which
\item \code{40} presbyterian church in the u.s.
\item \code{41} united presbyterian church in the u.s.a.
\item \code{42} other presbyterian churches
\item \code{43} presbyterian, merged
\item \code{48} presbyterian, don't know which
\item \code{50} episcopal church
\item \code{60} other, specify
\item \code{70} no denomination or non-denominational church
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
