% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_179.R
\name{discaffw}
\alias{discaffw}
\title{A woman won't get a job or promotion}
\source{
General Social Survey https://gss.norc.org
}
\description{
discaffw
}
\details{
Question 561b. What do you think the chances are these days that a woman won't get a job or promotion while an equally or less qualified man gets one instead?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab don't know \tab no answer \tab not very likely \tab somewhat likely \tab very likely \tab very unlikely \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2149 \tab 31 \tab 11 \tab 152 \tab 363 \tab 132 \tab 66 \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1889 \tab 29 \tab 129 \tab 153 \tab 357 \tab 200 \tab 60 \tab - \tab 2817 \cr
   2002 \tab 2295 \tab 21 \tab - \tab 65 \tab 234 \tab 121 \tab 29 \tab - \tab 2765 \cr
   2004 \tab 2359 \tab 17 \tab 1 \tab 65 \tab 218 \tab 116 \tab 36 \tab - \tab 2812 \cr
   2006 \tab 3512 \tab 39 \tab - \tab 198 \tab 440 \tab 236 \tab 85 \tab - \tab 4510 \cr
   2008 \tab 1342 \tab 9 \tab - \tab 135 \tab 323 \tab 165 \tab 49 \tab - \tab 2023 \cr
   2010 \tab 1345 \tab 25 \tab - \tab 127 \tab 312 \tab 170 \tab 65 \tab - \tab 2044 \cr
   2012 \tab 1332 \tab 21 \tab - \tab 133 \tab 294 \tab 134 \tab 60 \tab - \tab 1974 \cr
   2014 \tab 1689 \tab 19 \tab 1 \tab 164 \tab 413 \tab 175 \tab 77 \tab - \tab 2538 \cr
   2016 \tab 1915 \tab 17 \tab - \tab 163 \tab 480 \tab 211 \tab 81 \tab - \tab 2867 \cr
   2018 \tab 1583 \tab 22 \tab - \tab 139 \tab 338 \tab 197 \tab 69 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 2315 \tab 48 \tab - \tab 253 \tab 521 \tab 302 \tab 105 \tab - \tab 3544 \cr
   Total \tab 58937 \tab 298 \tab 142 \tab 1747 \tab 4293 \tab 2159 \tab 782 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} very unlikely
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
