% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.r
\name{.onAttach}
\alias{.onAttach}
\title{.onAttach}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{}

\item{pkgname}{}
}
\value{
Message
}
\description{
.onAttach
}
\details{
A message shown when attaching the package, to make clear what the
    main data objects are.
}
\author{
Kieran Healy
}
