% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_17.R
\name{dotpeop}
\alias{dotpeop}
\title{D.o.t. code - relation to people}
\source{
General Social Survey https://gss.norc.org
}
\description{
dotpeop
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 0.06 \tab 0.42 \tab 0.54 \tab 0.62 \tab 0.73 \tab 0.9 \tab 1.43 \tab 1.68 \tab 1.89 \tab 1.97 \tab 2 \tab 2.02 \tab 2.03 \tab 2.2 \tab 2.29 \tab 2.3 \tab 2.35 \tab 2.59 \tab 2.62 \tab 2.64 \tab 2.75 \tab 3.24 \tab 3.33 \tab 3.39 \tab 3.51 \tab 3.68 \tab 3.75 \tab 3.77 \tab 3.83 \tab 3.85 \tab 3.94 \tab 3.95 \tab 3.96 \tab 4 \tab 4.02 \tab 4.15 \tab 4.23 \tab 4.28 \tab 4.31 \tab 4.35 \tab 4.36 \tab 4.4 \tab 4.54 \tab 4.72 \tab 4.73 \tab 4.84 \tab 4.94 \tab 4.99 \tab 5.02 \tab 5.07 \tab 5.09 \tab 5.13 \tab 5.14 \tab 5.17 \tab 5.33 \tab 5.36 \tab 5.37 \tab 5.42 \tab 5.43 \tab 5.45 \tab 5.48 \tab 5.49 \tab 5.58 \tab 5.59 \tab 5.62 \tab 5.65 \tab 5.69 \tab 5.74 \tab 5.78 \tab 5.81 \tab 5.84 \tab 5.86 \tab 5.88 \tab 5.9 \tab 5.94 \tab 5.95 \tab 5.97 \tab 5.99 \tab 6.02 \tab 6.03 \tab 6.05 \tab 6.06 \tab 6.07 \tab 6.15 \tab 6.16 \tab 6.17 \tab 6.22 \tab 6.25 \tab 6.28 \tab 6.32 \tab 6.34 \tab 6.37 \tab 6.41 \tab 6.44 \tab 6.45 \tab 6.46 \tab 6.5 \tab 6.51 \tab 6.62 \tab 6.65 \tab 6.66 \tab 6.79 \tab 6.8 \tab 6.82 \tab 6.88 \tab 6.91 \tab 6.99 \tab 7 \tab 7.01 \tab 7.03 \tab 7.04 \tab 7.08 \tab 7.21 \tab 7.23 \tab 7.25 \tab 7.27 \tab 7.4 \tab 7.43 \tab 7.45 \tab 7.47 \tab 7.49 \tab 7.5 \tab 7.51 \tab 7.54 \tab 7.55 \tab 7.6 \tab 7.63 \tab 7.64 \tab 7.65 \tab 7.66 \tab 7.67 \tab 7.68 \tab 7.69 \tab 7.71 \tab 7.72 \tab 7.74 \tab 7.75 \tab 7.76 \tab 7.77 \tab 7.78 \tab 7.79 \tab 7.8 \tab 7.81 \tab 7.84 \tab 7.85 \tab 7.86 \tab 7.87 \tab 7.88 \tab 7.89 \tab 7.9 \tab 7.91 \tab 7.92 \tab 7.93 \tab 7.94 \tab 7.95 \tab 7.96 \tab 7.97 \tab 7.98 \tab 8 \tab iap \tab 0 \tab 1.44 \tab 2.09 \tab 2.47 \tab 2.49 \tab 3 \tab 3.06 \tab 3.42 \tab 3.97 \tab 4.04 \tab 4.39 \tab 4.7 \tab 5.01 \tab 5.28 \tab 5.57 \tab 5.63 \tab 5.67 \tab 6 \tab 6.29 \tab 6.47 \tab 6.49 \tab 6.98 \tab 7.57 \tab 7.58 \tab 7.73 \tab 7.82 \tab 2.44 \tab 5.18 \tab 5.5 \tab 5.75 \tab 6.04 \tab 6.09 \tab 6.94 \tab 7.05 \tab 7.14 \tab 7.16 \tab 7.22 \tab 7.7 \tab 3.66 \tab 4.08 \tab 5.23 \tab 5.27 \tab 5.38 \tab 6.08 \tab 6.93 \tab 7.02 \tab 7.24 \tab 7.28 \tab 2.39 \tab 4.33 \tab 4.86 \tab 5.26 \tab 7.07 \tab 7.36 \tab 2.51 \tab 3.55 \tab 4.1 \tab 5 \tab 7.83 \tab 4.93 \tab 7.12 \tab 7.15 \tab 7.59 \tab 2.82 \tab 5.06 \tab 5.29 \tab 4.17 \tab 5.8 \tab 6.55 \tab 7.62 \tab 7.11 \tab 3.48 \tab 5.66 \tab Total \cr
   1972 \tab 5 \tab 3 \tab 1 \tab 1 \tab 6 \tab 3 \tab 4 \tab 2 \tab 1 \tab 3 \tab 3 \tab 21 \tab 29 \tab 10 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 2 \tab 5 \tab 1 \tab 7 \tab 5 \tab 1 \tab 5 \tab 5 \tab 1 \tab 1 \tab 33 \tab 4 \tab 1 \tab 2 \tab 5 \tab 1 \tab 6 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 5 \tab 2 \tab 3 \tab 1 \tab 1 \tab 6 \tab 4 \tab 6 \tab 5 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 8 \tab 3 \tab 8 \tab 1 \tab 2 \tab 21 \tab 37 \tab 7 \tab 2 \tab 1 \tab 6 \tab 7 \tab 3 \tab 6 \tab 21 \tab 3 \tab 2 \tab 16 \tab 1 \tab 5 \tab 4 \tab 58 \tab 3 \tab 16 \tab 10 \tab 2 \tab 1 \tab 3 \tab 7 \tab 1 \tab 7 \tab 3 \tab 3 \tab 1 \tab 1 \tab 22 \tab 7 \tab 1 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 7 \tab 2 \tab 16 \tab 2 \tab 14 \tab 27 \tab 17 \tab 3 \tab 4 \tab 2 \tab 3 \tab 31 \tab 4 \tab 2 \tab 5 \tab 4 \tab 4 \tab 1 \tab 5 \tab 12 \tab 2 \tab 1 \tab 14 \tab 1 \tab 15 \tab 9 \tab 9 \tab 12 \tab 4 \tab 11 \tab 21 \tab 11 \tab 3 \tab 1 \tab 5 \tab 1 \tab 16 \tab 14 \tab 7 \tab 5 \tab 11 \tab 10 \tab 32 \tab 21 \tab 24 \tab 6 \tab 1 \tab 14 \tab 19 \tab 38 \tab 16 \tab 18 \tab 17 \tab 40 \tab 24 \tab 93 \tab 40 \tab 117 \tab 183 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 2 \tab 5 \tab 1 \tab 10 \tab - \tab 4 \tab 2 \tab - \tab - \tab 6 \tab 20 \tab 20 \tab 9 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 9 \tab 1 \tab 2 \tab 2 \tab - \tab 5 \tab 2 \tab 1 \tab 3 \tab 33 \tab 4 \tab 3 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 4 \tab 4 \tab - \tab 6 \tab 4 \tab 1 \tab 6 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 33 \tab 26 \tab 13 \tab - \tab - \tab 3 \tab 4 \tab 5 \tab 5 \tab 22 \tab 1 \tab 1 \tab 9 \tab - \tab 4 \tab 11 \tab 66 \tab 10 \tab 20 \tab 11 \tab 1 \tab 1 \tab 2 \tab 7 \tab - \tab 5 \tab 4 \tab 1 \tab - \tab 3 \tab 17 \tab 4 \tab 2 \tab 4 \tab 2 \tab - \tab 5 \tab 2 \tab 7 \tab 3 \tab 21 \tab 3 \tab 21 \tab 28 \tab 9 \tab 4 \tab 2 \tab - \tab - \tab 17 \tab 6 \tab - \tab 4 \tab 8 \tab 1 \tab 4 \tab 6 \tab 10 \tab 4 \tab 3 \tab 9 \tab 1 \tab 16 \tab 1 \tab 9 \tab 14 \tab 2 \tab 12 \tab 18 \tab 18 \tab 7 \tab 1 \tab 5 \tab 1 \tab 7 \tab 6 \tab 7 \tab 8 \tab 8 \tab 12 \tab 26 \tab 25 \tab 28 \tab 3 \tab 1 \tab 13 \tab 13 \tab 37 \tab 11 \tab 9 \tab 8 \tab 24 \tab 13 \tab 60 \tab 43 \tab 132 \tab 183 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab 4 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 5 \tab - \tab 3 \tab 2 \tab 5 \tab 1 \tab 10 \tab 2 \tab 2 \tab 5 \tab 8 \tab 26 \tab 44 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 8 \tab 2 \tab 7 \tab 6 \tab 1 \tab 2 \tab 8 \tab - \tab 4 \tab 19 \tab 7 \tab 1 \tab 4 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab 5 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 5 \tab 1 \tab 8 \tab 5 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 4 \tab 4 \tab 6 \tab 4 \tab - \tab 22 \tab 3 \tab 11 \tab 1 \tab 2 \tab 8 \tab 5 \tab 3 \tab 7 \tab 38 \tab 1 \tab 7 \tab 11 \tab - \tab 4 \tab 7 \tab 74 \tab 2 \tab 18 \tab 9 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 22 \tab 7 \tab - \tab 7 \tab 6 \tab - \tab 5 \tab 2 \tab 6 \tab 1 \tab 14 \tab 4 \tab 22 \tab 24 \tab 10 \tab 4 \tab 5 \tab 1 \tab - \tab 21 \tab - \tab 1 \tab 8 \tab 2 \tab 1 \tab 2 \tab 3 \tab 7 \tab 2 \tab 2 \tab 8 \tab 4 \tab 12 \tab 4 \tab 6 \tab 10 \tab 4 \tab 6 \tab 19 \tab 8 \tab 10 \tab 1 \tab 6 \tab 5 \tab 4 \tab 6 \tab 17 \tab 6 \tab 6 \tab 13 \tab 31 \tab 20 \tab 34 \tab 3 \tab 2 \tab 10 \tab 20 \tab 37 \tab 17 \tab 18 \tab 16 \tab 14 \tab 16 \tab 43 \tab 37 \tab 125 \tab 151 \tab 2 \tab - \tab 3 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 6 \tab 1 \tab 2 \tab - \tab 6 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 2 \tab - \tab 3 \tab - \tab 3 \tab 3 \tab 4 \tab - \tab - \tab 7 \tab 3 \tab 14 \tab 26 \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 10 \tab - \tab 3 \tab 4 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 26 \tab 6 \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 6 \tab 1 \tab 2 \tab - \tab 1 \tab 26 \tab 54 \tab 8 \tab 3 \tab 2 \tab 3 \tab 7 \tab 9 \tab 7 \tab 10 \tab 2 \tab 4 \tab 10 \tab 1 \tab 5 \tab 7 \tab 71 \tab - \tab 19 \tab 14 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 4 \tab - \tab - \tab - \tab 17 \tab 9 \tab 2 \tab 6 \tab 6 \tab - \tab 8 \tab 1 \tab 3 \tab 1 \tab 15 \tab 3 \tab 24 \tab 30 \tab 20 \tab 2 \tab 4 \tab 2 \tab - \tab 25 \tab 2 \tab 2 \tab 7 \tab 2 \tab 7 \tab 2 \tab 3 \tab 3 \tab 1 \tab 2 \tab 8 \tab 2 \tab 8 \tab 3 \tab 7 \tab 10 \tab 2 \tab 7 \tab 17 \tab 16 \tab 11 \tab - \tab 7 \tab 2 \tab 7 \tab 11 \tab 18 \tab 5 \tab 12 \tab 7 \tab 30 \tab 23 \tab 23 \tab 8 \tab 1 \tab 13 \tab 13 \tab 44 \tab 19 \tab 8 \tab 15 \tab 27 \tab 15 \tab 56 \tab 58 \tab 118 \tab 155 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 8 \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 7 \tab - \tab 4 \tab 2 \tab 6 \tab 1 \tab 3 \tab - \tab 2 \tab 6 \tab 2 \tab 26 \tab 29 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab 8 \tab 1 \tab - \tab 5 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 32 \tab 8 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 3 \tab 3 \tab - \tab - \tab - \tab 3 \tab 6 \tab 9 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab 18 \tab 63 \tab 12 \tab - \tab - \tab 6 \tab 7 \tab 3 \tab 6 \tab 16 \tab - \tab 6 \tab 12 \tab - \tab 1 \tab 8 \tab 64 \tab 5 \tab 23 \tab 13 \tab 1 \tab 1 \tab - \tab 5 \tab - \tab 8 \tab 6 \tab 1 \tab - \tab 1 \tab 17 \tab 8 \tab 1 \tab 6 \tab 2 \tab - \tab 1 \tab - \tab 8 \tab 3 \tab 9 \tab 4 \tab 16 \tab 25 \tab 9 \tab 1 \tab 3 \tab 2 \tab 3 \tab 21 \tab 6 \tab 1 \tab 5 \tab 6 \tab 4 \tab 1 \tab 8 \tab 8 \tab 2 \tab 3 \tab 9 \tab 3 \tab 13 \tab 4 \tab 9 \tab 11 \tab 1 \tab 8 \tab 22 \tab 13 \tab 5 \tab - \tab 2 \tab 6 \tab 10 \tab 6 \tab 5 \tab 3 \tab 13 \tab 7 \tab 17 \tab 15 \tab 34 \tab 6 \tab 3 \tab 15 \tab 24 \tab 53 \tab 9 \tab 15 \tab 18 \tab 24 \tab 10 \tab 72 \tab 50 \tab 112 \tab 158 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 9 \tab 3 \tab 1 \tab 2 \tab 3 \tab 1 \tab 5 \tab - \tab 1 \tab 1 \tab 7 \tab 28 \tab 35 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 5 \tab 4 \tab 3 \tab 4 \tab 1 \tab 1 \tab 6 \tab - \tab 1 \tab 28 \tab 3 \tab 1 \tab 4 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 5 \tab 1 \tab 3 \tab 2 \tab 2 \tab 10 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 20 \tab 58 \tab 5 \tab 1 \tab 2 \tab 2 \tab 7 \tab 7 \tab 1 \tab 10 \tab - \tab 2 \tab 18 \tab 1 \tab 7 \tab 8 \tab 68 \tab 7 \tab 26 \tab 6 \tab 2 \tab - \tab 3 \tab 2 \tab - \tab 4 \tab 3 \tab 1 \tab - \tab 2 \tab 19 \tab 15 \tab 3 \tab 4 \tab 4 \tab 1 \tab 1 \tab - \tab 9 \tab 3 \tab 15 \tab 5 \tab 20 \tab 30 \tab 10 \tab - \tab 7 \tab 1 \tab - \tab 21 \tab 1 \tab 1 \tab 4 \tab - \tab 6 \tab 2 \tab 5 \tab 7 \tab 1 \tab 1 \tab 6 \tab 4 \tab 9 \tab 5 \tab 16 \tab 8 \tab 2 \tab 8 \tab 15 \tab 18 \tab 6 \tab 1 \tab 9 \tab 6 \tab 5 \tab 11 \tab 6 \tab 4 \tab 12 \tab 4 \tab 34 \tab 21 \tab 39 \tab 13 \tab 1 \tab 17 \tab 8 \tab 33 \tab 13 \tab 16 \tab 21 \tab 21 \tab 9 \tab 60 \tab 55 \tab 165 \tab 133 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 3 \tab 2 \tab 3 \tab 4 \tab 8 \tab 1 \tab 2 \tab - \tab - \tab 5 \tab 2 \tab 17 \tab 49 \tab 3 \tab 2 \tab - \tab 1 \tab 5 \tab 2 \tab 3 \tab - \tab 2 \tab 11 \tab - \tab - \tab 6 \tab 1 \tab 3 \tab 6 \tab 1 \tab - \tab 21 \tab 3 \tab 2 \tab 5 \tab 1 \tab - \tab 8 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 9 \tab 5 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 22 \tab 63 \tab 10 \tab - \tab - \tab 3 \tab 1 \tab 10 \tab 6 \tab 11 \tab 1 \tab 1 \tab 17 \tab - \tab 4 \tab 7 \tab 80 \tab 7 \tab 16 \tab 11 \tab 2 \tab 1 \tab 2 \tab - \tab 3 \tab 5 \tab 4 \tab - \tab - \tab 1 \tab 18 \tab 11 \tab 1 \tab 2 \tab 2 \tab - \tab 3 \tab 1 \tab 3 \tab 1 \tab 20 \tab 1 \tab 19 \tab 35 \tab 14 \tab 1 \tab 10 \tab 1 \tab - \tab 21 \tab 2 \tab 2 \tab 4 \tab 6 \tab 1 \tab 2 \tab 3 \tab 3 \tab - \tab 4 \tab 10 \tab 2 \tab 11 \tab 1 \tab 7 \tab 10 \tab 1 \tab 8 \tab 17 \tab 15 \tab 5 \tab - \tab 7 \tab 2 \tab 8 \tab 19 \tab 7 \tab 4 \tab 18 \tab 9 \tab 25 \tab 18 \tab 31 \tab 4 \tab 2 \tab 18 \tab 16 \tab 52 \tab 14 \tab 22 \tab 16 \tab 26 \tab 20 \tab 48 \tab 46 \tab 128 \tab 149 \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 3 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab 8 \tab 3 \tab 6 \tab 1 \tab 5 \tab - \tab - \tab 2 \tab 2 \tab 17 \tab 28 \tab 6 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 5 \tab - \tab 3 \tab 8 \tab - \tab 2 \tab 7 \tab 1 \tab 2 \tab 25 \tab 5 \tab 1 \tab 6 \tab 1 \tab 3 \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 3 \tab 2 \tab 1 \tab 10 \tab 8 \tab 7 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 8 \tab - \tab 6 \tab - \tab 1 \tab 11 \tab 44 \tab 10 \tab 1 \tab 1 \tab 4 \tab 2 \tab 12 \tab 5 \tab 5 \tab - \tab 5 \tab 23 \tab 1 \tab 5 \tab 3 \tab 60 \tab 8 \tab 10 \tab 20 \tab 1 \tab - \tab 2 \tab 8 \tab - \tab 7 \tab 6 \tab - \tab 5 \tab - \tab 20 \tab 9 \tab 3 \tab 5 \tab 4 \tab 1 \tab 7 \tab - \tab 10 \tab 1 \tab 12 \tab 5 \tab 20 \tab 30 \tab 13 \tab 2 \tab 1 \tab 2 \tab - \tab 12 \tab 4 \tab 1 \tab 6 \tab 4 \tab 5 \tab 1 \tab 5 \tab 5 \tab 2 \tab 1 \tab 11 \tab - \tab 6 \tab 5 \tab 11 \tab 14 \tab 1 \tab 6 \tab 21 \tab 18 \tab 9 \tab - \tab 2 \tab 3 \tab 8 \tab 14 \tab 9 \tab 5 \tab 24 \tab 5 \tab 25 \tab 22 \tab 26 \tab 5 \tab 1 \tab 18 \tab 13 \tab 34 \tab 12 \tab 8 \tab 11 \tab 20 \tab 10 \tab 65 \tab 37 \tab 122 \tab 133 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 5 \tab 2 \tab 9 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 4 \tab 1 \tab 10 \tab 1 \tab 12 \tab 2 \tab 4 \tab - \tab - \tab 9 \tab 3 \tab 24 \tab 46 \tab 7 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 9 \tab 2 \tab 4 \tab 14 \tab 2 \tab 3 \tab 7 \tab 1 \tab 1 \tab 45 \tab 6 \tab 1 \tab 1 \tab - \tab 1 \tab 7 \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 5 \tab - \tab - \tab - \tab 11 \tab 59 \tab 9 \tab - \tab 2 \tab 4 \tab 6 \tab 4 \tab 5 \tab 11 \tab - \tab 1 \tab 11 \tab - \tab 7 \tab 12 \tab 64 \tab 5 \tab 23 \tab 10 \tab - \tab - \tab 4 \tab - \tab - \tab 10 \tab 6 \tab - \tab - \tab - \tab 25 \tab 7 \tab 2 \tab 8 \tab 10 \tab - \tab 3 \tab 1 \tab 16 \tab 6 \tab 24 \tab 16 \tab 42 \tab 28 \tab 21 \tab 4 \tab 3 \tab 2 \tab 3 \tab 51 \tab 3 \tab 1 \tab 8 \tab 2 \tab 12 \tab 3 \tab 3 \tab 14 \tab 3 \tab 4 \tab 41 \tab 1 \tab 18 \tab 9 \tab 27 \tab 24 \tab 1 \tab 13 \tab 14 \tab 9 \tab 10 \tab 1 \tab 2 \tab 6 \tab 6 \tab 16 \tab 10 \tab 8 \tab 18 \tab 12 \tab 25 \tab 21 \tab 39 \tab 7 \tab - \tab 24 \tab 17 \tab 50 \tab 3 \tab 24 \tab 22 \tab 30 \tab 17 \tab 51 \tab 66 \tab 164 \tab 148 \tab 3 \tab 1 \tab 4 \tab - \tab 3 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 5 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 5 \tab 4 \tab 4 \tab 2 \tab 12 \tab 2 \tab 2 \tab - \tab - \tab 9 \tab 5 \tab 22 \tab 33 \tab 5 \tab 1 \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 3 \tab 10 \tab - \tab 2 \tab 5 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 32 \tab 4 \tab - \tab 3 \tab 1 \tab 2 \tab 3 \tab - \tab 4 \tab - \tab - \tab 1 \tab 5 \tab 1 \tab 2 \tab 3 \tab 7 \tab 6 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 3 \tab 10 \tab 3 \tab 2 \tab - \tab 1 \tab 17 \tab 58 \tab 8 \tab 2 \tab 2 \tab 1 \tab 7 \tab 20 \tab 4 \tab 7 \tab - \tab 5 \tab 10 \tab - \tab 3 \tab 8 \tab 88 \tab 4 \tab 29 \tab 17 \tab 1 \tab - \tab 3 \tab 14 \tab 4 \tab 6 \tab 10 \tab - \tab 2 \tab 1 \tab 21 \tab 7 \tab - \tab 2 \tab 6 \tab - \tab 8 \tab - \tab 11 \tab 4 \tab 12 \tab 7 \tab 21 \tab 28 \tab 13 \tab 2 \tab 7 \tab 5 \tab 1 \tab 6 \tab 2 \tab 5 \tab 2 \tab 1 \tab 6 \tab 7 \tab 6 \tab 7 \tab - \tab 2 \tab 4 \tab - \tab 13 \tab 5 \tab 10 \tab 5 \tab 2 \tab 6 \tab 30 \tab 22 \tab 15 \tab 2 \tab 4 \tab 2 \tab 5 \tab 19 \tab 8 \tab 6 \tab 21 \tab 9 \tab 24 \tab 20 \tab 24 \tab 6 \tab 4 \tab 27 \tab 17 \tab 49 \tab 10 \tab 5 \tab 7 \tab 17 \tab 13 \tab 64 \tab 41 \tab 121 \tab 118 \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 6 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 9 \tab 1 \tab - \tab - \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 2 \tab - \tab 5 \tab 5 \tab 6 \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 23 \tab 25 \tab 6 \tab 1 \tab - \tab 1 \tab 7 \tab 1 \tab - \tab - \tab 2 \tab 7 \tab - \tab 5 \tab 6 \tab - \tab 3 \tab 5 \tab - \tab 2 \tab 32 \tab 4 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab - \tab 4 \tab 1 \tab 3 \tab 1 \tab 3 \tab 2 \tab 4 \tab 1 \tab 1 \tab 6 \tab - \tab 7 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 4 \tab 1 \tab 2 \tab - \tab 2 \tab 21 \tab 56 \tab 2 \tab - \tab 2 \tab 5 \tab 9 \tab 11 \tab 6 \tab 9 \tab 1 \tab 5 \tab 22 \tab 1 \tab 6 \tab 6 \tab 53 \tab 7 \tab 26 \tab 13 \tab - \tab - \tab 3 \tab 13 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 22 \tab 5 \tab 1 \tab 5 \tab 7 \tab - \tab 3 \tab - \tab 5 \tab 3 \tab 41 \tab 5 \tab 17 \tab 25 \tab 21 \tab 3 \tab 11 \tab 2 \tab 1 \tab 6 \tab 3 \tab 1 \tab 2 \tab 3 \tab 2 \tab 16 \tab 2 \tab 3 \tab - \tab - \tab 17 \tab - \tab 9 \tab 3 \tab 2 \tab 12 \tab 4 \tab 8 \tab 27 \tab 20 \tab 15 \tab - \tab 3 \tab 3 \tab 14 \tab 12 \tab 13 \tab 2 \tab 23 \tab 6 \tab 24 \tab 21 \tab 32 \tab 3 \tab 1 \tab 21 \tab 8 \tab 33 \tab 12 \tab 4 \tab 7 \tab 18 \tab 15 \tab 59 \tab 35 \tab 99 \tab 113 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 5 \tab 1 \tab 3 \tab 1 \tab 5 \tab 3 \tab 2 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1473 \cr
   1985 \tab 2 \tab 6 \tab 8 \tab 6 \tab 3 \tab 1 \tab 4 \tab - \tab 4 \tab 7 \tab 6 \tab 14 \tab 28 \tab 3 \tab 2 \tab - \tab - \tab 9 \tab 2 \tab 4 \tab - \tab 1 \tab 13 \tab 1 \tab 2 \tab 11 \tab 1 \tab 3 \tab 6 \tab - \tab - \tab 21 \tab 6 \tab 3 \tab 3 \tab 4 \tab 3 \tab 4 \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 3 \tab 6 \tab 2 \tab 1 \tab 3 \tab 9 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab 4 \tab 4 \tab 2 \tab 7 \tab - \tab 4 \tab 10 \tab 46 \tab 10 \tab 1 \tab 1 \tab 4 \tab 7 \tab 12 \tab 8 \tab 8 \tab 3 \tab 4 \tab 21 \tab - \tab 7 \tab 6 \tab 60 \tab 6 \tab 18 \tab 18 \tab 2 \tab - \tab 4 \tab 2 \tab - \tab 4 \tab 5 \tab - \tab 7 \tab 3 \tab 22 \tab 1 \tab 3 \tab 4 \tab 9 \tab 2 \tab 3 \tab - \tab 9 \tab 8 \tab 17 \tab 1 \tab 22 \tab 37 \tab 13 \tab 2 \tab 7 \tab 1 \tab - \tab 11 \tab 3 \tab 2 \tab 8 \tab 4 \tab 10 \tab 1 \tab 5 \tab 9 \tab 3 \tab 2 \tab 23 \tab 1 \tab 13 \tab 3 \tab 14 \tab 13 \tab 5 \tab 4 \tab 21 \tab 22 \tab 8 \tab 1 \tab 4 \tab 3 \tab 8 \tab 21 \tab 10 \tab 2 \tab 26 \tab 5 \tab 15 \tab 18 \tab 22 \tab 5 \tab 2 \tab 29 \tab 14 \tab 47 \tab 9 \tab 7 \tab 19 \tab 28 \tab 12 \tab 55 \tab 48 \tab 89 \tab 85 \tab 2 \tab - \tab 2 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 7 \tab - \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 5 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1 \tab 1 \tab 3 \tab 6 \tab 5 \tab 5 \tab 5 \tab - \tab 2 \tab 1 \tab 9 \tab 16 \tab 25 \tab 3 \tab 2 \tab - \tab 2 \tab 2 \tab 5 \tab 3 \tab - \tab 2 \tab 8 \tab 2 \tab 2 \tab 16 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 40 \tab 8 \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 3 \tab 3 \tab 2 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 6 \tab 8 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 6 \tab 1 \tab 5 \tab - \tab 3 \tab 18 \tab 65 \tab 8 \tab - \tab 1 \tab 5 \tab 6 \tab 7 \tab 8 \tab 11 \tab - \tab 3 \tab 18 \tab 1 \tab 8 \tab 15 \tab 55 \tab 5 \tab 31 \tab 16 \tab 2 \tab - \tab 1 \tab 14 \tab 1 \tab 1 \tab 8 \tab - \tab 6 \tab 1 \tab 30 \tab 3 \tab - \tab 2 \tab 5 \tab - \tab 11 \tab 1 \tab 5 \tab 4 \tab 18 \tab 1 \tab 20 \tab 23 \tab 15 \tab 6 \tab 5 \tab 3 \tab - \tab 1 \tab 4 \tab 1 \tab 7 \tab 1 \tab 3 \tab 10 \tab 4 \tab 7 \tab 4 \tab 1 \tab 5 \tab 2 \tab 10 \tab 2 \tab 11 \tab 7 \tab 1 \tab 15 \tab 19 \tab 17 \tab 10 \tab 2 \tab 4 \tab 1 \tab 10 \tab 4 \tab 9 \tab 3 \tab 22 \tab 11 \tab 20 \tab 14 \tab 21 \tab 7 \tab 1 \tab 20 \tab 17 \tab 27 \tab 19 \tab 3 \tab 6 \tab 21 \tab 9 \tab 62 \tab 22 \tab 105 \tab 109 \tab 7 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 1 \tab 1 \tab 2 \tab 3 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 6 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 2 \tab 1 \tab 6 \tab 5 \tab 12 \tab 3 \tab 3 \tab 1 \tab 2 \tab 9 \tab 5 \tab 20 \tab 26 \tab 13 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab 6 \tab - \tab - \tab 8 \tab - \tab 4 \tab 15 \tab 1 \tab - \tab 3 \tab - \tab 2 \tab 60 \tab 5 \tab 3 \tab 2 \tab 2 \tab 4 \tab 3 \tab 1 \tab 7 \tab 3 \tab 4 \tab 5 \tab 2 \tab 3 \tab 7 \tab 1 \tab 14 \tab 6 \tab 4 \tab 2 \tab - \tab - \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab 2 \tab 3 \tab 3 \tab - \tab 1 \tab 22 \tab 59 \tab 8 \tab 1 \tab 2 \tab 1 \tab 10 \tab 7 \tab 4 \tab 11 \tab 1 \tab 2 \tab 14 \tab - \tab 5 \tab 14 \tab 67 \tab 3 \tab 28 \tab 16 \tab - \tab - \tab 2 \tab 23 \tab 1 \tab 9 \tab 2 \tab 3 \tab 2 \tab 2 \tab 26 \tab 11 \tab 4 \tab 5 \tab 11 \tab 1 \tab 20 \tab 1 \tab 16 \tab 8 \tab 20 \tab 4 \tab 41 \tab 11 \tab 27 \tab 1 \tab 1 \tab 3 \tab 1 \tab 11 \tab 3 \tab - \tab 6 \tab 3 \tab 9 \tab 10 \tab 2 \tab 7 \tab 2 \tab 1 \tab 9 \tab 5 \tab 9 \tab 8 \tab 29 \tab 11 \tab 3 \tab 13 \tab 28 \tab 40 \tab 18 \tab - \tab 2 \tab 4 \tab 12 \tab 12 \tab 21 \tab 7 \tab 32 \tab 8 \tab 29 \tab 34 \tab 22 \tab 6 \tab 2 \tab 23 \tab 13 \tab 27 \tab 14 \tab 4 \tab 13 \tab 20 \tab 19 \tab 61 \tab 50 \tab 117 \tab 113 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 9 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab 4 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab 7 \tab 1 \tab 3 \tab 4 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1819 \cr
   1988 \tab 5 \tab 2 \tab 5 \tab 5 \tab 10 \tab 2 \tab 3 \tab - \tab 3 \tab 5 \tab 8 \tab 2 \tab 26 \tab 3 \tab 7 \tab - \tab 1 \tab 4 \tab 6 \tab 3 \tab - \tab 4 \tab 11 \tab - \tab 5 \tab 7 \tab 2 \tab 2 \tab 6 \tab 1 \tab 1 \tab 33 \tab 3 \tab 4 \tab 1 \tab - \tab 11 \tab 2 \tab - \tab 3 \tab 2 \tab 2 \tab 4 \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab 5 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 3 \tab 3 \tab 3 \tab 5 \tab - \tab - \tab 11 \tab 69 \tab 6 \tab 1 \tab 3 \tab 9 \tab 10 \tab 3 \tab 5 \tab 11 \tab 4 \tab 5 \tab 14 \tab - \tab 6 \tab 8 \tab 60 \tab 5 \tab 29 \tab 17 \tab 2 \tab 1 \tab 5 \tab 12 \tab - \tab 1 \tab 4 \tab - \tab 3 \tab 1 \tab 21 \tab 6 \tab 4 \tab 4 \tab 15 \tab - \tab 5 \tab - \tab 5 \tab 4 \tab 9 \tab 4 \tab 20 \tab 17 \tab 26 \tab 2 \tab 3 \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab 4 \tab 5 \tab 2 \tab 5 \tab - \tab 7 \tab - \tab 3 \tab 22 \tab - \tab 6 \tab 1 \tab 18 \tab 14 \tab 2 \tab 11 \tab 35 \tab 22 \tab 15 \tab 3 \tab 2 \tab 4 \tab 1 \tab 14 \tab 12 \tab 5 \tab 26 \tab 4 \tab 29 \tab 11 \tab 27 \tab 5 \tab 5 \tab 13 \tab 10 \tab 33 \tab 11 \tab 3 \tab 10 \tab 17 \tab 18 \tab 42 \tab 45 \tab 97 \tab 93 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1481 \cr
   1989 \tab 2 \tab 1 \tab 16 \tab 3 \tab 6 \tab 5 \tab 4 \tab - \tab 1 \tab 10 \tab 9 \tab 16 \tab 26 \tab 4 \tab 7 \tab - \tab 2 \tab 5 \tab - \tab 2 \tab - \tab 4 \tab 13 \tab 1 \tab 7 \tab 12 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab 35 \tab 2 \tab 3 \tab 1 \tab 1 \tab 7 \tab 3 \tab 2 \tab 5 \tab 2 \tab 5 \tab 3 \tab 1 \tab 6 \tab 6 \tab - \tab 3 \tab 6 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 9 \tab 2 \tab 6 \tab - \tab 3 \tab 11 \tab 63 \tab 5 \tab - \tab 4 \tab 6 \tab 8 \tab 13 \tab 8 \tab 6 \tab - \tab 5 \tab 16 \tab - \tab 4 \tab 8 \tab 68 \tab 5 \tab 30 \tab 16 \tab - \tab - \tab 6 \tab 2 \tab 3 \tab 2 \tab 7 \tab 2 \tab 7 \tab 2 \tab 29 \tab 8 \tab 5 \tab 5 \tab 11 \tab - \tab 4 \tab - \tab 8 \tab 10 \tab 11 \tab 1 \tab 28 \tab 22 \tab 15 \tab 1 \tab 5 \tab 1 \tab - \tab 9 \tab 1 \tab 2 \tab 6 \tab 2 \tab 3 \tab 1 \tab 2 \tab 9 \tab 3 \tab 2 \tab 16 \tab 2 \tab 14 \tab 2 \tab 13 \tab 16 \tab 7 \tab 8 \tab 17 \tab 17 \tab 15 \tab - \tab 7 \tab 4 \tab 5 \tab 16 \tab 14 \tab 5 \tab 14 \tab 5 \tab 31 \tab 22 \tab 29 \tab 8 \tab 3 \tab 20 \tab 13 \tab 26 \tab 9 \tab 8 \tab 15 \tab 23 \tab 13 \tab 42 \tab 27 \tab 86 \tab 106 \tab 1 \tab 1 \tab 5 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1537 \cr
   1990 \tab 2 \tab 1 \tab 8 \tab 4 \tab 12 \tab 2 \tab 6 \tab - \tab 4 \tab 8 \tab 10 \tab 10 \tab 29 \tab 6 \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 8 \tab - \tab 1 \tab 9 \tab 2 \tab 1 \tab 12 \tab - \tab 2 \tab 4 \tab - \tab 4 \tab 29 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 3 \tab 2 \tab 3 \tab 14 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 4 \tab - \tab - \tab 18 \tab 52 \tab 8 \tab - \tab - \tab 3 \tab 4 \tab 12 \tab 2 \tab 13 \tab - \tab 5 \tab 28 \tab - \tab 3 \tab 9 \tab 58 \tab 9 \tab 21 \tab 15 \tab - \tab - \tab 4 \tab 6 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 24 \tab 3 \tab 4 \tab 5 \tab 9 \tab 1 \tab 8 \tab - \tab 5 \tab 10 \tab 16 \tab 3 \tab 19 \tab 16 \tab 14 \tab 2 \tab 3 \tab 2 \tab - \tab 6 \tab 2 \tab 2 \tab 9 \tab 2 \tab 8 \tab 3 \tab 3 \tab 6 \tab 2 \tab - \tab 10 \tab 1 \tab 18 \tab 9 \tab 11 \tab 9 \tab 1 \tab 15 \tab 20 \tab 21 \tab 11 \tab - \tab 7 \tab 3 \tab 9 \tab 6 \tab 9 \tab 7 \tab 12 \tab 4 \tab 21 \tab 18 \tab 18 \tab 3 \tab 2 \tab 14 \tab 2 \tab 26 \tab 8 \tab 18 \tab 12 \tab 30 \tab 14 \tab 41 \tab 32 \tab 90 \tab 80 \tab 1 \tab 2 \tab 5 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 4 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 56 \tab 27 \tab 93 \tab 52 \tab 125 \tab 36 \tab 69 \tab 7 \tab 23 \tab 90 \tab 90 \tab 316 \tab 524 \tab 88 \tab 36 \tab 4 \tab 20 \tab 44 \tab 25 \tab 46 \tab 7 \tab 27 \tab 149 \tab 17 \tab 57 \tab 138 \tab 16 \tab 36 \tab 75 \tab 10 \tab 28 \tab 544 \tab 81 \tab 35 \tab 38 \tab 24 \tab 37 \tab 53 \tab 8 \tab 52 \tab 29 \tab 23 \tab 29 \tab 47 \tab 27 \tab 46 \tab 18 \tab 49 \tab 116 \tab 51 \tab 41 \tab 19 \tab 12 \tab 9 \tab 11 \tab 27 \tab 7 \tab 28 \tab 76 \tab 33 \tab 64 \tab 7 \tab 24 \tab 312 \tab 875 \tab 140 \tab 13 \tab 25 \tab 73 \tab 107 \tab 141 \tab 93 \tab 220 \tab 17 \tab 63 \tab 270 \tab 6 \tab 84 \tab 141 \tab 1114 \tab 91 \tab 383 \tab 232 \tab 20 \tab 6 \tab 47 \tab 119 \tab 15 \tab 80 \tab 77 \tab 14 \tab 35 \tab 22 \tab 372 \tab 121 \tab 36 \tab 78 \tab 112 \tab 7 \tab 97 \tab 10 \tab 133 \tab 72 \tab 290 \tab 69 \tab 386 \tab 436 \tab 267 \tab 40 \tab 81 \tab 31 \tab 13 \tab 275 \tab 47 \tab 24 \tab 95 \tab 55 \tab 84 \tab 71 \tab 65 \tab 124 \tab 31 \tab 32 \tab 222 \tab 29 \tab 200 \tab 74 \tab 209 \tab 200 \tab 43 \tab 159 \tab 361 \tab 307 \tab 173 \tab 13 \tab 78 \tab 56 \tab 135 \tab 207 \tab 182 \tab 85 \tab 298 \tab 131 \tab 438 \tab 344 \tab 473 \tab 98 \tab 32 \tab 309 \tab 237 \tab 646 \tab 206 \tab 190 \tab 233 \tab 400 \tab 247 \tab 974 \tab 732 \tab 1987 \tab 48335 \tab 25 \tab 13 \tab 27 \tab 5 \tab 22 \tab 17 \tab 23 \tab 8 \tab 5 \tab 41 \tab 6 \tab 16 \tab 28 \tab 13 \tab 21 \tab 15 \tab 7 \tab 25 \tab 16 \tab 39 \tab 22 \tab 56 \tab 7 \tab 45 \tab 13 \tab 19 \tab 8 \tab 15 \tab 8 \tab 27 \tab 22 \tab 18 \tab 28 \tab 16 \tab 4 \tab 12 \tab 14 \tab 18 \tab 16 \tab 2 \tab 5 \tab 4 \tab 19 \tab 17 \tab 21 \tab 4 \tab 3 \tab 1 \tab 6 \tab 8 \tab 11 \tab 13 \tab 3 \tab 7 \tab 2 \tab 6 \tab 7 \tab 2 \tab 5 \tab 9 \tab 18 \tab 4 \tab 2 \tab 2 \tab 1 \tab 4 \tab 13 \tab 10 \tab 1 \tab 7 \tab 1 \tab 3 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
