% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_18.R
\name{dotpres}
\alias{dotpres}
\title{dotpres}
\source{
General Social Survey https://gss.norc.org
}
\description{
Temme prestige scale
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 0 \tab 10.8 \tab 11.4 \tab 11.5 \tab 13.8 \tab 14.1 \tab 14.5 \tab 14.6 \tab 15.3 \tab 15.5 \tab 15.8 \tab 16.7 \tab 16.8 \tab 16.9 \tab 17.2 \tab 17.4 \tab 17.8 \tab 18.2 \tab 18.3 \tab 18.7 \tab 19.2 \tab 19.9 \tab 2 \tab 20.2 \tab 20.4 \tab 20.5 \tab 20.9 \tab 21.3 \tab 21.4 \tab 21.6 \tab 21.7 \tab 22 \tab 22.6 \tab 22.7 \tab 22.9 \tab 23 \tab 23.3 \tab 23.4 \tab 23.5 \tab 23.6 \tab 23.9 \tab 24.1 \tab 24.2 \tab 24.4 \tab 24.8 \tab 25 \tab 25.2 \tab 25.3 \tab 25.6 \tab 25.9 \tab 26 \tab 26.3 \tab 26.4 \tab 26.7 \tab 26.8 \tab 26.9 \tab 27 \tab 27.4 \tab 28 \tab 28.3 \tab 28.5 \tab 28.7 \tab 28.8 \tab 28.9 \tab 29 \tab 29.1 \tab 29.2 \tab 29.3 \tab 29.6 \tab 29.7 \tab 29.8 \tab 29.9 \tab 30.2 \tab 30.3 \tab 30.5 \tab 30.7 \tab 30.8 \tab 30.9 \tab 31.2 \tab 31.3 \tab 31.7 \tab 31.8 \tab 32 \tab 32.1 \tab 32.2 \tab 32.3 \tab 32.4 \tab 33 \tab 33.1 \tab 33.2 \tab 33.4 \tab 33.5 \tab 33.6 \tab 33.7 \tab 34 \tab 34.1 \tab 34.4 \tab 34.5 \tab 34.9 \tab 35.2 \tab 35.3 \tab 35.4 \tab 35.5 \tab 35.6 \tab 35.7 \tab 35.8 \tab 36.2 \tab 36.3 \tab 36.6 \tab 37.3 \tab 37.5 \tab 37.6 \tab 37.7 \tab 37.9 \tab 38.2 \tab 38.4 \tab 38.5 \tab 38.6 \tab 38.7 \tab 38.8 \tab 38.9 \tab 39 \tab 39.2 \tab 39.4 \tab 39.5 \tab 39.7 \tab 39.8 \tab 39.9 \tab 40.5 \tab 40.6 \tab 40.7 \tab 40.8 \tab 41 \tab 41.2 \tab 41.4 \tab 41.5 \tab 41.6 \tab 41.7 \tab 42.4 \tab 42.5 \tab 42.6 \tab 42.8 \tab 43.2 \tab 43.3 \tab 43.4 \tab 43.6 \tab 43.7 \tab 43.8 \tab 43.9 \tab 44.1 \tab 44.2 \tab 44.4 \tab 44.5 \tab 44.6 \tab 44.7 \tab 44.9 \tab 45.2 \tab 45.3 \tab 45.4 \tab 45.5 \tab 45.7 \tab 45.9 \tab 46.3 \tab 46.4 \tab 46.7 \tab 46.9 \tab 47.1 \tab 47.5 \tab 47.6 \tab 47.7 \tab 47.8 \tab 47.9 \tab 48 \tab 48.3 \tab 48.6 \tab 49 \tab 49.1 \tab 49.5 \tab 49.7 \tab 50.2 \tab 50.4 \tab 50.7 \tab 50.9 \tab 51.8 \tab 51.9 \tab 52.2 \tab 52.3 \tab 52.4 \tab 52.5 \tab 52.8 \tab 53.7 \tab 53.8 \tab 54.2 \tab 55.3 \tab 55.5 \tab 55.8 \tab 56 \tab 56.3 \tab 56.4 \tab 57.3 \tab 58.3 \tab 59.5 \tab 59.6 \tab 60.1 \tab 60.5 \tab 60.6 \tab 60.7 \tab 60.8 \tab 60.9 \tab 61.3 \tab 61.7 \tab 62 \tab 62.5 \tab 62.6 \tab 62.7 \tab 63.6 \tab 63.7 \tab 64.1 \tab 64.6 \tab 65.3 \tab 65.5 \tab 66 \tab 67.1 \tab 67.4 \tab 67.9 \tab 68.1 \tab 68.2 \tab 68.5 \tab 68.8 \tab 69 \tab 69.3 \tab 69.7 \tab 70.3 \tab 70.4 \tab 70.9 \tab 71 \tab 71.2 \tab 71.9 \tab 73 \tab 76.4 \tab 76.7 \tab 8.5 \tab 88.4 \tab 9.3 \tab 9.8 \tab iap \tab 12.8 \tab 15 \tab 18.1 \tab 19.3 \tab 23.1 \tab 23.2 \tab 24.3 \tab 24.7 \tab 24.9 \tab 25.1 \tab 25.7 \tab 25.8 \tab 26.5 \tab 26.6 \tab 30.4 \tab 30.6 \tab 31 \tab 32.6 \tab 32.7 \tab 33.9 \tab 34.2 \tab 34.8 \tab 35.1 \tab 39.1 \tab 40.4 \tab 41.3 \tab 42.7 \tab 43 \tab 45 \tab 46 \tab 48.4 \tab 49.4 \tab 49.6 \tab 5.3 \tab 53.5 \tab 53.9 \tab 58 \tab 61.8 \tab 64.5 \tab 64.7 \tab 67 \tab 7.4 \tab 71.8 \tab 72.1 \tab 72.8 \tab 73.2 \tab 73.3 \tab 75.3 \tab 12.6 \tab 19 \tab 2.1 \tab 21 \tab 27.3 \tab 27.5 \tab 27.6 \tab 33.3 \tab 36 \tab 38.1 \tab 39.6 \tab 42.9 \tab 43.5 \tab 45.6 \tab 48.1 \tab 68.9 \tab 74.9 \tab 78 \tab 25.5 \tab 27.9 \tab 37.4 \tab 40.9 \tab 42.1 \tab 43.1 \tab 47.2 \tab 61.4 \tab 72 \tab 74.7 \tab 46.8 \tab 48.8 \tab 65.2 \tab 67.6 \tab 68 \tab 71.1 \tab 73.7 \tab 27.1 \tab 30 \tab 30.1 \tab 35 \tab 46.5 \tab 49.3 \tab 62.9 \tab 26.2 \tab 56.1 \tab 59.7 \tab 19.8 \tab 27.2 \tab 32.8 \tab 36.7 \tab 39.3 \tab 45.8 \tab 54.6 \tab 58.5 \tab 68.4 \tab 77.1 \tab 28.6 \tab 61.9 \tab 15.7 \tab 21.1 \tab 46.2 \tab 55.7 \tab 18.4 \tab 20.6 \tab 28.4 \tab 3.7 \tab 24 \tab 73.4 \tab 22.2 \tab 46.1 \tab 73.6 \tab 74.2 \tab 71.3 \tab 36.4 \tab 37.2 \tab Total \cr
   1972 \tab 1 \tab 13 \tab 31 \tab 3 \tab 6 \tab 2 \tab 3 \tab 4 \tab 3 \tab 3 \tab 7 \tab 2 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 13 \tab 7 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 2 \tab 11 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 2 \tab 12 \tab 18 \tab 2 \tab 14 \tab 2 \tab 14 \tab 28 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 3 \tab 1 \tab 4 \tab 5 \tab 3 \tab 1 \tab 4 \tab 1 \tab 6 \tab 2 \tab 1 \tab 7 \tab 1 \tab 23 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 18 \tab 1 \tab 43 \tab 3 \tab 6 \tab 4 \tab 3 \tab 1 \tab 6 \tab 21 \tab 20 \tab 7 \tab 7 \tab 4 \tab 40 \tab 10 \tab 6 \tab 1 \tab 7 \tab 1 \tab 2 \tab 1 \tab 7 \tab 2 \tab 6 \tab 1 \tab 3 \tab 11 \tab 2 \tab 15 \tab 1 \tab 3 \tab 2 \tab 16 \tab 9 \tab 9 \tab 9 \tab 1 \tab 1 \tab 3 \tab 13 \tab 12 \tab 1 \tab 6 \tab 4 \tab 6 \tab 6 \tab 38 \tab 18 \tab 2 \tab 2 \tab 1 \tab 3 \tab 13 \tab 2 \tab 15 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 10 \tab 5 \tab 2 \tab 4 \tab 1 \tab 2 \tab 12 \tab 13 \tab 1 \tab 16 \tab 9 \tab 1 \tab 12 \tab 6 \tab 1 \tab 2 \tab 2 \tab 10 \tab 15 \tab 9 \tab 17 \tab 9 \tab 3 \tab 1 \tab 1 \tab 8 \tab 1 \tab 4 \tab 7 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 27 \tab 2 \tab 8 \tab 2 \tab 1 \tab 12 \tab 2 \tab 1 \tab 57 \tab 6 \tab 3 \tab 5 \tab 24 \tab 3 \tab 5 \tab 11 \tab 5 \tab 7 \tab 1 \tab 3 \tab 6 \tab 7 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 3 \tab 2 \tab 3 \tab 22 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 4 \tab 1 \tab 4 \tab 5 \tab 1 \tab 5 \tab 37 \tab 4 \tab 9 \tab 6 \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab 3 \tab 22 \tab 29 \tab 1 \tab 2 \tab 3 \tab 8 \tab 2 \tab 3 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 4 \tab 183 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 4 \tab 7 \tab 17 \tab 5 \tab 8 \tab - \tab 1 \tab 2 \tab 4 \tab 4 \tab 7 \tab - \tab 4 \tab 5 \tab 1 \tab - \tab - \tab 9 \tab 6 \tab - \tab 6 \tab 1 \tab - \tab 1 \tab 1 \tab 12 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 3 \tab 5 \tab 22 \tab 2 \tab 10 \tab 5 \tab 14 \tab 28 \tab 2 \tab 1 \tab - \tab 9 \tab - \tab - \tab - \tab 4 \tab 3 \tab - \tab - \tab 10 \tab - \tab 5 \tab 2 \tab - \tab 9 \tab - \tab 26 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 17 \tab - \tab 23 \tab - \tab 5 \tab 3 \tab 1 \tab 1 \tab 1 \tab 18 \tab 18 \tab 4 \tab 16 \tab 5 \tab 24 \tab 2 \tab 7 \tab 6 \tab 5 \tab 2 \tab 1 \tab 1 \tab 12 \tab 1 \tab 8 \tab 3 \tab 4 \tab 8 \tab - \tab 29 \tab 1 \tab 4 \tab 3 \tab 7 \tab 4 \tab 4 \tab 10 \tab 1 \tab - \tab 2 \tab 8 \tab 7 \tab - \tab 2 \tab 7 \tab 6 \tab 5 \tab 27 \tab 9 \tab 3 \tab - \tab - \tab 4 \tab 19 \tab 1 \tab 20 \tab - \tab 5 \tab - \tab - \tab 1 \tab 2 \tab 6 \tab 6 \tab 6 \tab 1 \tab 11 \tab 1 \tab 2 \tab 15 \tab 15 \tab 3 \tab 18 \tab 1 \tab 2 \tab 8 \tab 3 \tab 1 \tab 1 \tab 6 \tab 4 \tab 14 \tab 5 \tab 10 \tab 3 \tab 4 \tab 3 \tab - \tab 2 \tab 2 \tab 4 \tab 4 \tab - \tab 5 \tab - \tab - \tab 1 \tab 1 \tab 31 \tab 3 \tab 5 \tab 1 \tab - \tab 13 \tab 3 \tab 1 \tab 63 \tab 4 \tab 7 \tab 2 \tab 35 \tab 2 \tab 11 \tab 9 \tab 1 \tab 13 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab - \tab 2 \tab 17 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 6 \tab 2 \tab 2 \tab - \tab 40 \tab 1 \tab 12 \tab 9 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 20 \tab 20 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 5 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab 183 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab 2 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 9 \tab 7 \tab 21 \tab 1 \tab 3 \tab - \tab 1 \tab 5 \tab 7 \tab 5 \tab 3 \tab 1 \tab 3 \tab 5 \tab 1 \tab - \tab - \tab 10 \tab 1 \tab - \tab 7 \tab - \tab - \tab 1 \tab 3 \tab 10 \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 16 \tab 6 \tab 18 \tab - \tab 7 \tab 25 \tab 1 \tab - \tab - \tab 8 \tab - \tab 2 \tab - \tab 5 \tab 2 \tab - \tab - \tab 7 \tab 1 \tab 5 \tab - \tab 2 \tab 4 \tab 1 \tab 25 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 17 \tab 1 \tab 18 \tab 1 \tab 6 \tab 5 \tab 2 \tab 1 \tab 2 \tab 26 \tab 23 \tab 7 \tab 24 \tab 3 \tab 21 \tab 7 \tab 5 \tab 4 \tab 9 \tab 1 \tab 2 \tab 5 \tab 11 \tab 1 \tab 7 \tab 2 \tab 1 \tab 5 \tab 3 \tab 29 \tab 3 \tab 3 \tab 4 \tab 17 \tab 4 \tab 4 \tab 3 \tab 3 \tab - \tab 4 \tab 8 \tab 13 \tab - \tab 3 \tab 5 \tab 10 \tab 7 \tab 5 \tab 16 \tab 4 \tab - \tab 2 \tab - \tab 15 \tab 2 \tab 21 \tab - \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 9 \tab 3 \tab - \tab 4 \tab 1 \tab 3 \tab 16 \tab 10 \tab 1 \tab 12 \tab 4 \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 3 \tab 8 \tab 21 \tab 9 \tab 11 \tab 5 \tab 4 \tab 4 \tab 2 \tab 6 \tab 2 \tab 7 \tab 5 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 32 \tab - \tab 5 \tab 1 \tab 1 \tab 13 \tab 3 \tab 1 \tab 67 \tab 3 \tab 3 \tab 6 \tab 28 \tab - \tab 2 \tab - \tab 5 \tab 11 \tab 1 \tab 5 \tab 1 \tab 7 \tab - \tab 1 \tab 1 \tab 8 \tab 3 \tab 5 \tab 1 \tab 4 \tab 22 \tab 1 \tab 1 \tab 3 \tab 1 \tab 5 \tab 3 \tab 2 \tab 3 \tab 8 \tab 2 \tab 5 \tab 27 \tab 2 \tab 4 \tab 5 \tab 7 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab - \tab 27 \tab 44 \tab 5 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 10 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 151 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 3 \tab 7 \tab 25 \tab 1 \tab 7 \tab - \tab 2 \tab 4 \tab 6 \tab 3 \tab 1 \tab 2 \tab 1 \tab 8 \tab 4 \tab - \tab - \tab 8 \tab 3 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 8 \tab 1 \tab - \tab - \tab 4 \tab 6 \tab 2 \tab 2 \tab 18 \tab 3 \tab 8 \tab 1 \tab 10 \tab 31 \tab 1 \tab 1 \tab - \tab 8 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab 2 \tab - \tab 7 \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab 25 \tab 6 \tab - \tab 1 \tab 1 \tab - \tab 38 \tab 1 \tab 27 \tab - \tab 9 \tab 3 \tab 1 \tab 1 \tab 4 \tab 19 \tab 21 \tab 9 \tab - \tab 1 \tab 24 \tab 5 \tab 8 \tab 2 \tab 6 \tab 1 \tab 1 \tab 5 \tab 11 \tab 1 \tab 4 \tab - \tab 3 \tab 7 \tab 1 \tab 34 \tab - \tab 4 \tab 4 \tab 7 \tab 8 \tab 4 \tab 4 \tab 1 \tab - \tab 6 \tab 4 \tab 7 \tab - \tab 3 \tab 11 \tab 14 \tab 7 \tab 58 \tab 19 \tab 5 \tab - \tab 1 \tab 2 \tab 21 \tab 2 \tab 22 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 2 \tab - \tab 9 \tab 3 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 9 \tab 1 \tab 20 \tab 3 \tab 1 \tab 4 \tab 9 \tab 2 \tab 1 \tab 2 \tab 14 \tab 11 \tab 8 \tab 11 \tab 8 \tab 2 \tab 1 \tab 2 \tab 1 \tab 4 \tab 6 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 33 \tab 1 \tab 4 \tab 2 \tab 1 \tab 5 \tab - \tab 2 \tab 68 \tab 4 \tab 3 \tab 5 \tab 32 \tab - \tab 3 \tab 4 \tab 5 \tab 9 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab 3 \tab 5 \tab 3 \tab - \tab 1 \tab 17 \tab 1 \tab - \tab 4 \tab - \tab 4 \tab - \tab 2 \tab 9 \tab - \tab 1 \tab 2 \tab 35 \tab 2 \tab 11 \tab 3 \tab 7 \tab 3 \tab 1 \tab - \tab 2 \tab 5 \tab 1 \tab 17 \tab 26 \tab 2 \tab 1 \tab 3 \tab 2 \tab 2 \tab 4 \tab 3 \tab - \tab 7 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab 2 \tab 6 \tab 155 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 6 \tab 1 \tab - \tab - \tab 9 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 3 \tab 11 \tab 21 \tab 2 \tab 10 \tab - \tab - \tab 3 \tab 8 \tab 2 \tab 5 \tab 1 \tab 1 \tab 3 \tab 3 \tab 3 \tab - \tab 10 \tab 6 \tab - \tab 7 \tab - \tab 2 \tab - \tab 2 \tab 8 \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 13 \tab 4 \tab 18 \tab 4 \tab 9 \tab 26 \tab - \tab 1 \tab - \tab 6 \tab - \tab 3 \tab - \tab 9 \tab 2 \tab - \tab - \tab 6 \tab - \tab 2 \tab 2 \tab - \tab 4 \tab - \tab 28 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 32 \tab - \tab 24 \tab 1 \tab 4 \tab 5 \tab 1 \tab 2 \tab 2 \tab 25 \tab 27 \tab 8 \tab - \tab 2 \tab 38 \tab 7 \tab 6 \tab 1 \tab 6 \tab - \tab - \tab 5 \tab 14 \tab - \tab 3 \tab 1 \tab 4 \tab 14 \tab - \tab 19 \tab 1 \tab 2 \tab 4 \tab 14 \tab 3 \tab 5 \tab 4 \tab 1 \tab 1 \tab 4 \tab 8 \tab 14 \tab - \tab 1 \tab 9 \tab 15 \tab 7 \tab 64 \tab 21 \tab 1 \tab - \tab 1 \tab 7 \tab 9 \tab 1 \tab 7 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 11 \tab 8 \tab 2 \tab 7 \tab - \tab 2 \tab 5 \tab 14 \tab 2 \tab 9 \tab 4 \tab - \tab 6 \tab 4 \tab 1 \tab 1 \tab 3 \tab 9 \tab 8 \tab 8 \tab 8 \tab 3 \tab 1 \tab 2 \tab - \tab 5 \tab 4 \tab 8 \tab 7 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 37 \tab 3 \tab 7 \tab 3 \tab 1 \tab 7 \tab 1 \tab - \tab 61 \tab 6 \tab 6 \tab 5 \tab 20 \tab 1 \tab 5 \tab 5 \tab 3 \tab 12 \tab 2 \tab 4 \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 6 \tab 2 \tab 4 \tab - \tab - \tab 17 \tab - \tab - \tab 1 \tab - \tab 4 \tab 3 \tab 3 \tab 3 \tab 2 \tab - \tab 7 \tab 40 \tab - \tab 9 \tab 6 \tab 6 \tab - \tab 3 \tab - \tab 1 \tab 4 \tab 2 \tab 27 \tab 29 \tab 3 \tab 2 \tab 1 \tab 3 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab 4 \tab - \tab - \tab 2 \tab 1 \tab - \tab 158 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 7 \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 3 \tab 10 \tab 21 \tab 3 \tab 12 \tab - \tab 2 \tab 7 \tab 6 \tab 3 \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab 5 \tab 3 \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 18 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 16 \tab 4 \tab 18 \tab 1 \tab 22 \tab 31 \tab 1 \tab 1 \tab - \tab 9 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 8 \tab 1 \tab 9 \tab 3 \tab 4 \tab 4 \tab - \tab 32 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 28 \tab 1 \tab 25 \tab 1 \tab 5 \tab 6 \tab - \tab - \tab 1 \tab 17 \tab 30 \tab 15 \tab - \tab 1 \tab 25 \tab 14 \tab 13 \tab 3 \tab 3 \tab 3 \tab 1 \tab 4 \tab 8 \tab - \tab 3 \tab 2 \tab 2 \tab 14 \tab - \tab 27 \tab - \tab 6 \tab 4 \tab 14 \tab 5 \tab 2 \tab 3 \tab 2 \tab 1 \tab 3 \tab 8 \tab 10 \tab - \tab 3 \tab 4 \tab 23 \tab 1 \tab 61 \tab 11 \tab 1 \tab - \tab 2 \tab 7 \tab 17 \tab 2 \tab 22 \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 10 \tab 4 \tab - \tab 2 \tab 1 \tab 2 \tab 8 \tab 4 \tab 2 \tab 9 \tab 5 \tab - \tab 3 \tab 6 \tab - \tab - \tab 5 \tab 7 \tab 4 \tab 10 \tab 17 \tab 7 \tab - \tab 3 \tab 2 \tab 4 \tab 5 \tab 3 \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 23 \tab - \tab 7 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab 66 \tab 10 \tab 4 \tab 4 \tab 24 \tab 2 \tab 6 \tab 5 \tab 5 \tab 6 \tab 1 \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 19 \tab 4 \tab - \tab 5 \tab 2 \tab 3 \tab 3 \tab - \tab 7 \tab 6 \tab - \tab 9 \tab 33 \tab 2 \tab 11 \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab 4 \tab 2 \tab 3 \tab 28 \tab 35 \tab 2 \tab 5 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 4 \tab 3 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 3 \tab 133 \tab 3 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 7 \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 6 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 3 \tab 10 \tab 21 \tab 2 \tab 7 \tab - \tab 2 \tab 10 \tab 5 \tab 9 \tab - \tab 2 \tab - \tab 5 \tab - \tab - \tab - \tab 10 \tab 3 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 2 \tab - \tab 1 \tab 6 \tab 3 \tab 4 \tab 12 \tab 5 \tab 14 \tab - \tab 11 \tab 35 \tab - \tab 1 \tab - \tab 12 \tab - \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 2 \tab - \tab 1 \tab 5 \tab - \tab 18 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 31 \tab - \tab 21 \tab - \tab 4 \tab 2 \tab 2 \tab 2 \tab 1 \tab 24 \tab 20 \tab 11 \tab - \tab 4 \tab 16 \tab 5 \tab 4 \tab 5 \tab 6 \tab - \tab 4 \tab 4 \tab 10 \tab 1 \tab 5 \tab 1 \tab 2 \tab 8 \tab 1 \tab 24 \tab 2 \tab 4 \tab 3 \tab 20 \tab 3 \tab 5 \tab 5 \tab - \tab - \tab - \tab 6 \tab 10 \tab - \tab 3 \tab 5 \tab 9 \tab 6 \tab 64 \tab 14 \tab 2 \tab - \tab 1 \tab 4 \tab 19 \tab 2 \tab 17 \tab 4 \tab 5 \tab - \tab - \tab 3 \tab - \tab - \tab 11 \tab 3 \tab 2 \tab 4 \tab 1 \tab 2 \tab 10 \tab 11 \tab 2 \tab 12 \tab 1 \tab - \tab 5 \tab 2 \tab 2 \tab 3 \tab 5 \tab 10 \tab 6 \tab 4 \tab 17 \tab 12 \tab 1 \tab - \tab - \tab 8 \tab 2 \tab 3 \tab 8 \tab 2 \tab 3 \tab 3 \tab - \tab 4 \tab 1 \tab 45 \tab - \tab 9 \tab 4 \tab 1 \tab 7 \tab 1 \tab - \tab 72 \tab 9 \tab 3 \tab 6 \tab 24 \tab 2 \tab 6 \tab 4 \tab 3 \tab 11 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab 6 \tab 2 \tab - \tab 2 \tab 18 \tab 2 \tab 3 \tab 2 \tab - \tab 4 \tab 3 \tab - \tab 10 \tab 6 \tab 1 \tab 3 \tab 32 \tab 5 \tab 8 \tab 8 \tab 5 \tab 5 \tab 3 \tab 1 \tab 4 \tab 5 \tab 3 \tab 17 \tab 49 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 4 \tab - \tab 2 \tab - \tab - \tab 3 \tab 2 \tab - \tab 4 \tab 1 \tab 4 \tab 149 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 6 \tab 1 \tab 3 \tab - \tab 4 \tab - \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 3 \tab 8 \tab 12 \tab 1 \tab 7 \tab 1 \tab - \tab 1 \tab 3 \tab 5 \tab 8 \tab 1 \tab 2 \tab 4 \tab - \tab - \tab - \tab 12 \tab 4 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 8 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 21 \tab 8 \tab 11 \tab 3 \tab 8 \tab 30 \tab 1 \tab 1 \tab 1 \tab 6 \tab - \tab - \tab - \tab 8 \tab 1 \tab - \tab - \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab 3 \tab - \tab 15 \tab 4 \tab 1 \tab - \tab - \tab - \tab 25 \tab 1 \tab 18 \tab 1 \tab 3 \tab 5 \tab 1 \tab 3 \tab 1 \tab 20 \tab 17 \tab 9 \tab - \tab 3 \tab 25 \tab 4 \tab 5 \tab 1 \tab 9 \tab 1 \tab 2 \tab 8 \tab 5 \tab - \tab 4 \tab - \tab 1 \tab 12 \tab - \tab 26 \tab - \tab 1 \tab 1 \tab 8 \tab 2 \tab 2 \tab 11 \tab 1 \tab - \tab 5 \tab 5 \tab 5 \tab - \tab 3 \tab 9 \tab 11 \tab 5 \tab 46 \tab 15 \tab 2 \tab - \tab 5 \tab 9 \tab 23 \tab 1 \tab 12 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 7 \tab 5 \tab 1 \tab 11 \tab 2 \tab - \tab 5 \tab 11 \tab 2 \tab 21 \tab 5 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 5 \tab 9 \tab 4 \tab 7 \tab 18 \tab 10 \tab 2 \tab 2 \tab 1 \tab 7 \tab 9 \tab 5 \tab 4 \tab 3 \tab 2 \tab 2 \tab 5 \tab 1 \tab 1 \tab 25 \tab 1 \tab 12 \tab 5 \tab - \tab 10 \tab 3 \tab 1 \tab 56 \tab 10 \tab 3 \tab 8 \tab 15 \tab 2 \tab 8 \tab 8 \tab 3 \tab 11 \tab 1 \tab 6 \tab 7 \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 3 \tab - \tab - \tab 18 \tab 4 \tab - \tab 3 \tab 1 \tab 4 \tab 3 \tab 2 \tab 13 \tab 7 \tab - \tab - \tab 30 \tab 1 \tab 15 \tab 6 \tab 2 \tab 1 \tab 2 \tab - \tab 3 \tab 12 \tab 1 \tab 20 \tab 28 \tab 3 \tab 5 \tab 1 \tab 8 \tab 4 \tab 4 \tab - \tab 3 \tab 3 \tab 2 \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab - \tab - \tab 8 \tab - \tab - \tab 3 \tab - \tab 3 \tab 133 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 13 \tab 1 \tab - \tab - \tab 9 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 5 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 4 \tab 12 \tab 51 \tab 3 \tab 24 \tab - \tab 4 \tab 3 \tab 10 \tab 6 \tab - \tab 1 \tab 4 \tab 7 \tab 1 \tab 3 \tab - \tab 41 \tab 7 \tab - \tab 13 \tab - \tab - \tab 1 \tab 4 \tab 19 \tab - \tab 1 \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 13 \tab 2 \tab 29 \tab 2 \tab 19 \tab 28 \tab - \tab - \tab - \tab 14 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 12 \tab - \tab 10 \tab 1 \tab 1 \tab 6 \tab - \tab 33 \tab 10 \tab - \tab - \tab - \tab - \tab 24 \tab 1 \tab 30 \tab 2 \tab 1 \tab 6 \tab - \tab 3 \tab 1 \tab 18 \tab 22 \tab 7 \tab - \tab 3 \tab 22 \tab 9 \tab 7 \tab 6 \tab 4 \tab - \tab 6 \tab 6 \tab 7 \tab - \tab 5 \tab - \tab 3 \tab 9 \tab - \tab 50 \tab 3 \tab 4 \tab 2 \tab 21 \tab 5 \tab 6 \tab 14 \tab 4 \tab - \tab 5 \tab 17 \tab 11 \tab - \tab 7 \tab 2 \tab 18 \tab 5 \tab 62 \tab 22 \tab 1 \tab - \tab 5 \tab 11 \tab 25 \tab - \tab 15 \tab 1 \tab 6 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 12 \tab 2 \tab 1 \tab 3 \tab - \tab 2 \tab 16 \tab 10 \tab 11 \tab 6 \tab 9 \tab 2 \tab 6 \tab 1 \tab 1 \tab 2 \tab 9 \tab 12 \tab 10 \tab 21 \tab 17 \tab 13 \tab 4 \tab 1 \tab 2 \tab 15 \tab 3 \tab 6 \tab 3 \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 34 \tab - \tab 3 \tab 3 \tab 2 \tab 11 \tab 1 \tab - \tab 62 \tab 4 \tab 8 \tab 15 \tab 21 \tab 5 \tab 3 \tab 7 \tab - \tab 9 \tab - \tab 3 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 4 \tab 7 \tab 6 \tab 1 \tab 5 \tab 25 \tab 3 \tab - \tab 2 \tab - \tab - \tab 4 \tab 1 \tab 4 \tab 7 \tab 1 \tab 4 \tab 52 \tab 5 \tab 5 \tab 12 \tab 9 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 24 \tab 46 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 4 \tab - \tab - \tab 10 \tab 1 \tab - \tab 1 \tab - \tab 8 \tab 148 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 6 \tab 5 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 4 \tab 2 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 3 \tab 7 \tab 6 \tab 2 \tab 9 \tab - \tab 1 \tab 7 \tab 1 \tab 8 \tab 14 \tab 5 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 5 \tab 7 \tab - \tab 6 \tab 1 \tab - \tab 1 \tab - \tab 10 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 15 \tab 1 \tab 16 \tab - \tab 10 \tab 29 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 8 \tab - \tab 12 \tab 2 \tab 2 \tab 4 \tab - \tab 33 \tab 6 \tab 2 \tab 1 \tab - \tab - \tab 23 \tab - \tab 24 \tab - \tab 1 \tab 5 \tab 2 \tab - \tab 2 \tab 30 \tab 25 \tab 7 \tab - \tab - \tab 35 \tab 2 \tab 7 \tab 7 \tab 7 \tab - \tab 2 \tab 4 \tab 11 \tab - \tab 4 \tab 1 \tab 2 \tab 10 \tab 1 \tab 33 \tab 2 \tab - \tab 1 \tab 9 \tab 2 \tab 5 \tab 4 \tab - \tab - \tab 8 \tab 5 \tab 9 \tab 1 \tab 1 \tab 3 \tab 16 \tab 5 \tab 60 \tab 17 \tab 1 \tab - \tab - \tab 9 \tab 24 \tab 1 \tab 14 \tab - \tab 7 \tab - \tab - \tab 1 \tab 2 \tab - \tab 5 \tab 6 \tab 2 \tab 2 \tab 2 \tab 1 \tab 10 \tab 4 \tab 3 \tab 16 \tab 5 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 3 \tab 8 \tab 10 \tab 7 \tab 7 \tab 12 \tab 2 \tab 3 \tab 2 \tab 12 \tab 9 \tab 4 \tab 8 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 40 \tab 1 \tab 11 \tab 8 \tab 1 \tab 16 \tab 1 \tab - \tab 87 \tab 6 \tab 4 \tab 5 \tab 23 \tab - \tab 6 \tab 8 \tab 5 \tab 8 \tab 3 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 3 \tab 21 \tab 3 \tab 4 \tab 3 \tab 2 \tab 4 \tab 4 \tab 2 \tab 21 \tab 2 \tab - \tab 5 \tab 41 \tab 3 \tab 18 \tab 12 \tab 9 \tab 2 \tab 3 \tab - \tab 2 \tab 6 \tab 4 \tab 25 \tab 33 \tab 4 \tab 1 \tab 2 \tab 3 \tab 3 \tab 5 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab 7 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 4 \tab - \tab 2 \tab - \tab 4 \tab 118 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 5 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 5 \tab 9 \tab 6 \tab - \tab 2 \tab - \tab 9 \tab 11 \tab 9 \tab 3 \tab 13 \tab 1 \tab 8 \tab 8 \tab 2 \tab 1 \tab - \tab 17 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab 6 \tab - \tab 1 \tab - \tab - \tab 6 \tab 1 \tab 7 \tab 13 \tab 4 \tab 13 \tab 3 \tab 10 \tab 27 \tab - \tab - \tab - \tab 10 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 28 \tab 3 \tab - \tab - \tab - \tab - \tab 18 \tab - \tab 24 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 26 \tab 19 \tab 5 \tab - \tab - \tab 21 \tab 2 \tab 3 \tab 6 \tab 2 \tab - \tab - \tab - \tab 16 \tab - \tab 28 \tab - \tab - \tab 10 \tab - \tab 30 \tab 1 \tab - \tab - \tab 7 \tab 1 \tab 11 \tab 3 \tab - \tab - \tab 1 \tab 5 \tab 7 \tab 1 \tab 2 \tab 8 \tab 12 \tab 6 \tab 58 \tab 6 \tab 5 \tab - \tab 4 \tab 13 \tab 21 \tab - \tab 18 \tab - \tab 17 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 2 \tab 8 \tab 2 \tab 12 \tab 3 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 12 \tab 9 \tab 7 \tab 19 \tab 5 \tab 3 \tab 2 \tab 2 \tab 12 \tab 5 \tab 4 \tab 7 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 26 \tab 2 \tab 2 \tab 1 \tab - \tab 6 \tab 1 \tab 2 \tab 52 \tab 6 \tab 5 \tab 6 \tab 28 \tab 1 \tab 7 \tab 8 \tab 3 \tab 3 \tab 1 \tab 1 \tab 7 \tab 5 \tab - \tab 6 \tab - \tab 5 \tab 6 \tab 1 \tab 2 \tab 1 \tab 22 \tab 1 \tab 1 \tab 1 \tab 3 \tab 4 \tab 4 \tab 5 \tab 11 \tab 5 \tab - \tab 2 \tab 38 \tab 2 \tab 16 \tab 6 \tab 3 \tab 7 \tab - \tab 3 \tab 6 \tab 7 \tab 4 \tab 26 \tab 25 \tab 1 \tab 1 \tab 3 \tab 5 \tab 3 \tab 7 \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 5 \tab - \tab - \tab 5 \tab 3 \tab 9 \tab 113 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 5 \tab 2 \tab - \tab 4 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 3 \tab 12 \tab 11 \tab - \tab 9 \tab - \tab - \tab 7 \tab 11 \tab 10 \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab - \tab 23 \tab 2 \tab - \tab 6 \tab 2 \tab - \tab - \tab - \tab 14 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 5 \tab 13 \tab 1 \tab 12 \tab 1 \tab 10 \tab 38 \tab 1 \tab 1 \tab - \tab 6 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 6 \tab - \tab 3 \tab 1 \tab 1 \tab 4 \tab - \tab 23 \tab 5 \tab 2 \tab 1 \tab - \tab - \tab 20 \tab - \tab 31 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 21 \tab 13 \tab 1 \tab - \tab 3 \tab 26 \tab 6 \tab 6 \tab 2 \tab 11 \tab 1 \tab - \tab 1 \tab 5 \tab 1 \tab 4 \tab 3 \tab 3 \tab 8 \tab 1 \tab 28 \tab 1 \tab 1 \tab 2 \tab 5 \tab 4 \tab 7 \tab 2 \tab - \tab - \tab 7 \tab 10 \tab 7 \tab - \tab 2 \tab 12 \tab 13 \tab 8 \tab 50 \tab 19 \tab 1 \tab - \tab 3 \tab 6 \tab 19 \tab 2 \tab 10 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 13 \tab 5 \tab 2 \tab 4 \tab - \tab - \tab 10 \tab 12 \tab 7 \tab 8 \tab 3 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 3 \tab 7 \tab 8 \tab 3 \tab 18 \tab 19 \tab 2 \tab - \tab 1 \tab 21 \tab 9 \tab 6 \tab 3 \tab 1 \tab 3 \tab 1 \tab 7 \tab 1 \tab - \tab 38 \tab 4 \tab 6 \tab 4 \tab 1 \tab 7 \tab 1 \tab 1 \tab 62 \tab 9 \tab 1 \tab 11 \tab 19 \tab 1 \tab 5 \tab 8 \tab 3 \tab 15 \tab 4 \tab 3 \tab 3 \tab 2 \tab - \tab 3 \tab - \tab 4 \tab 5 \tab 4 \tab 6 \tab 2 \tab 21 \tab 3 \tab - \tab 4 \tab 1 \tab 2 \tab 7 \tab 2 \tab 13 \tab 6 \tab - \tab 2 \tab 33 \tab 4 \tab 14 \tab 3 \tab 7 \tab 9 \tab 4 \tab - \tab 2 \tab 4 \tab 6 \tab 18 \tab 28 \tab 2 \tab 7 \tab 1 \tab 6 \tab 2 \tab 6 \tab 2 \tab - \tab 7 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab - \tab 4 \tab 8 \tab 6 \tab - \tab 6 \tab - \tab 6 \tab 85 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 4 \tab 1 \tab 7 \tab - \tab 2 \tab - \tab 2 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab 2 \tab 4 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 5 \tab 1 \tab 1 \tab 8 \tab - \tab 3 \tab 5 \tab 4 \tab 9 \tab 14 \tab 1 \tab 2 \tab 4 \tab 2 \tab - \tab - \tab 7 \tab 1 \tab - \tab 7 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 11 \tab 16 \tab 8 \tab 3 \tab 8 \tab 24 \tab 1 \tab 1 \tab - \tab 8 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 13 \tab - \tab 4 \tab 1 \tab 1 \tab 2 \tab - \tab 34 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 12 \tab 1 \tab 25 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 21 \tab 16 \tab 3 \tab - \tab 6 \tab 29 \tab 2 \tab 7 \tab 2 \tab 8 \tab - \tab 1 \tab 2 \tab 7 \tab - \tab 2 \tab 1 \tab 3 \tab 11 \tab - \tab 29 \tab - \tab 1 \tab 1 \tab 4 \tab 3 \tab 5 \tab 3 \tab 6 \tab - \tab 2 \tab 8 \tab 9 \tab 1 \tab 3 \tab 10 \tab 14 \tab 8 \tab 72 \tab 14 \tab 2 \tab - \tab 1 \tab 6 \tab 30 \tab 2 \tab 16 \tab - \tab 10 \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab 4 \tab 1 \tab 6 \tab 1 \tab 3 \tab 6 \tab 5 \tab 2 \tab 15 \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab - \tab 7 \tab 9 \tab 9 \tab 4 \tab 22 \tab 2 \tab 6 \tab 4 \tab 1 \tab 13 \tab 5 \tab 8 \tab 3 \tab 1 \tab - \tab 2 \tab 6 \tab - \tab - \tab 16 \tab 3 \tab 7 \tab 4 \tab 3 \tab 9 \tab 1 \tab 1 \tab 52 \tab 8 \tab 11 \tab 17 \tab 23 \tab 3 \tab 5 \tab 6 \tab 2 \tab 11 \tab 2 \tab 3 \tab 4 \tab 2 \tab - \tab 5 \tab - \tab 5 \tab 3 \tab 5 \tab 1 \tab 1 \tab 30 \tab 1 \tab 1 \tab 3 \tab 2 \tab 4 \tab 1 \tab - \tab 8 \tab 2 \tab 1 \tab 1 \tab 48 \tab 2 \tab 12 \tab 5 \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab 4 \tab 3 \tab 19 \tab 25 \tab 3 \tab 4 \tab 5 \tab 6 \tab 2 \tab 5 \tab 2 \tab 2 \tab 4 \tab 3 \tab - \tab - \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 5 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 6 \tab 1 \tab 6 \tab 109 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 6 \tab 1 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 5 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 4 \tab 11 \tab 11 \tab - \tab 29 \tab - \tab 1 \tab 2 \tab 4 \tab 6 \tab 23 \tab - \tab 7 \tab 11 \tab 4 \tab 1 \tab 1 \tab 10 \tab 5 \tab - \tab 7 \tab - \tab - \tab 3 \tab - \tab 26 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 22 \tab 6 \tab 9 \tab 1 \tab 10 \tab 11 \tab 3 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 4 \tab 9 \tab 1 \tab - \tab 14 \tab - \tab 13 \tab 1 \tab - \tab 11 \tab - \tab 30 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 24 \tab - \tab 27 \tab - \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 27 \tab 23 \tab 11 \tab - \tab 3 \tab 32 \tab 1 \tab 6 \tab 3 \tab 10 \tab - \tab - \tab 1 \tab 7 \tab - \tab 1 \tab 2 \tab 7 \tab 11 \tab - \tab 57 \tab 1 \tab 2 \tab 1 \tab 3 \tab 7 \tab 8 \tab 7 \tab 2 \tab 2 \tab 2 \tab 7 \tab 7 \tab 1 \tab 2 \tab 13 \tab 20 \tab 4 \tab 62 \tab 16 \tab 1 \tab - \tab 2 \tab 12 \tab 41 \tab - \tab 20 \tab 1 \tab 10 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 9 \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab 11 \tab 10 \tab - \tab 9 \tab 8 \tab - \tab 4 \tab - \tab - \tab 1 \tab 4 \tab 15 \tab 9 \tab 8 \tab 16 \tab 12 \tab 1 \tab 3 \tab 2 \tab 18 \tab 5 \tab 5 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 16 \tab 1 \tab 10 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 70 \tab 6 \tab 10 \tab 15 \tab 33 \tab 3 \tab 4 \tab 15 \tab 2 \tab 11 \tab 4 \tab 2 \tab 2 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab 9 \tab 4 \tab 3 \tab 3 \tab 26 \tab 1 \tab 1 \tab 2 \tab 5 \tab 8 \tab 6 \tab - \tab 9 \tab 3 \tab 1 \tab 2 \tab 68 \tab 2 \tab 33 \tab 12 \tab 10 \tab 1 \tab 6 \tab 4 \tab 1 \tab 5 \tab 3 \tab 21 \tab 26 \tab 2 \tab 1 \tab 3 \tab 7 \tab 1 \tab 10 \tab 3 \tab 4 \tab 6 \tab 4 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 2 \tab 6 \tab 1 \tab 1 \tab 5 \tab - \tab 11 \tab 113 \tab 3 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab 5 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab 9 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 4 \tab 4 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 2 \tab 2 \tab 5 \tab 1 \tab 13 \tab - \tab 1 \tab 3 \tab 12 \tab 1 \tab 12 \tab - \tab - \tab 5 \tab 2 \tab 1 \tab - \tab 22 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab 14 \tab - \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 5 \tab 6 \tab 2 \tab 6 \tab 1 \tab 6 \tab 17 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 8 \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab 29 \tab 2 \tab 4 \tab 2 \tab - \tab - \tab 28 \tab 1 \tab 25 \tab - \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 37 \tab 21 \tab 6 \tab - \tab - \tab 20 \tab 4 \tab 5 \tab 1 \tab 6 \tab - \tab - \tab 3 \tab 11 \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab - \tab 31 \tab - \tab 5 \tab 4 \tab 6 \tab 9 \tab 2 \tab 3 \tab 1 \tab - \tab 3 \tab 9 \tab 10 \tab - \tab 6 \tab 12 \tab 20 \tab 5 \tab 72 \tab 8 \tab 3 \tab 1 \tab 3 \tab 3 \tab 15 \tab 2 \tab 17 \tab 3 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab 3 \tab 3 \tab 1 \tab 2 \tab 5 \tab 8 \tab 3 \tab 10 \tab 1 \tab 2 \tab 5 \tab 1 \tab - \tab 1 \tab 3 \tab 14 \tab 8 \tab 8 \tab 12 \tab 8 \tab 2 \tab 1 \tab 3 \tab 13 \tab 6 \tab 3 \tab 8 \tab 2 \tab 2 \tab 2 \tab 3 \tab 3 \tab 1 \tab 28 \tab - \tab 2 \tab 3 \tab 2 \tab 6 \tab - \tab 1 \tab 56 \tab 5 \tab 4 \tab 7 \tab 26 \tab 2 \tab 6 \tab 8 \tab - \tab 7 \tab 3 \tab 7 \tab 3 \tab 5 \tab - \tab 6 \tab 1 \tab 9 \tab 5 \tab 2 \tab 3 \tab 3 \tab 20 \tab 1 \tab - \tab 2 \tab 2 \tab 3 \tab 9 \tab 3 \tab 3 \tab 6 \tab - \tab 5 \tab 42 \tab 1 \tab 19 \tab 10 \tab 5 \tab 4 \tab 3 \tab 2 \tab 2 \tab 5 \tab 4 \tab 5 \tab 26 \tab 2 \tab 3 \tab 2 \tab 12 \tab 2 \tab 10 \tab 3 \tab 1 \tab 6 \tab 9 \tab 4 \tab 2 \tab 6 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 3 \tab 5 \tab 2 \tab - \tab 5 \tab - \tab 12 \tab 93 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 5 \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1481 \cr
   1989 \tab 5 \tab 3 \tab 9 \tab 1 \tab 9 \tab - \tab 2 \tab 5 \tab 9 \tab 6 \tab 2 \tab 2 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab 16 \tab 1 \tab - \tab 10 \tab - \tab - \tab - \tab - \tab 12 \tab - \tab - \tab - \tab 1 \tab 4 \tab 2 \tab 2 \tab 14 \tab 2 \tab 13 \tab 1 \tab 4 \tab 23 \tab - \tab - \tab - \tab 5 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 2 \tab 8 \tab - \tab 3 \tab 1 \tab - \tab 5 \tab - \tab 31 \tab 5 \tab 1 \tab 2 \tab 1 \tab - \tab 10 \tab - \tab 28 \tab - \tab 8 \tab 1 \tab 1 \tab 2 \tab - \tab 22 \tab 17 \tab 8 \tab - \tab 4 \tab 18 \tab 2 \tab 9 \tab 2 \tab 5 \tab - \tab - \tab 1 \tab 10 \tab - \tab 5 \tab 2 \tab 4 \tab 12 \tab - \tab 38 \tab 3 \tab 1 \tab 2 \tab 8 \tab 2 \tab 4 \tab 7 \tab 3 \tab 2 \tab 1 \tab 10 \tab 5 \tab - \tab 4 \tab 12 \tab 13 \tab 9 \tab 64 \tab 7 \tab 2 \tab - \tab 1 \tab 5 \tab 14 \tab - \tab 16 \tab 1 \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab 12 \tab 2 \tab - \tab 8 \tab 2 \tab 1 \tab 12 \tab 9 \tab 2 \tab 14 \tab 2 \tab - \tab 5 \tab - \tab - \tab 1 \tab 3 \tab 13 \tab 11 \tab 7 \tab 15 \tab 14 \tab 1 \tab 2 \tab 4 \tab 8 \tab 4 \tab 2 \tab 4 \tab 1 \tab - \tab 1 \tab 7 \tab 2 \tab 1 \tab 17 \tab 3 \tab 6 \tab 5 \tab - \tab 6 \tab 1 \tab - \tab 63 \tab 6 \tab 2 \tab 13 \tab 22 \tab 2 \tab 3 \tab 6 \tab 1 \tab 6 \tab 2 \tab 3 \tab 2 \tab 7 \tab 2 \tab 1 \tab - \tab 6 \tab 8 \tab 4 \tab 6 \tab 2 \tab 28 \tab 3 \tab 3 \tab 3 \tab 6 \tab 5 \tab 9 \tab 1 \tab 14 \tab 4 \tab - \tab 2 \tab 48 \tab 2 \tab 14 \tab 6 \tab 10 \tab 5 \tab 2 \tab 5 \tab 2 \tab 4 \tab 7 \tab 16 \tab 26 \tab 4 \tab 4 \tab 5 \tab 13 \tab 3 \tab 8 \tab 3 \tab 2 \tab 5 \tab 7 \tab 5 \tab 1 \tab 5 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab 16 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab 106 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 6 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 6 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1537 \cr
   1990 \tab 4 \tab 6 \tab 6 \tab - \tab 4 \tab - \tab 1 \tab 3 \tab 4 \tab 5 \tab 6 \tab 2 \tab 2 \tab 2 \tab 10 \tab - \tab - \tab 9 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 13 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 6 \tab 13 \tab 2 \tab 12 \tab 2 \tab 11 \tab 16 \tab - \tab 2 \tab - \tab 4 \tab - \tab 1 \tab - \tab 3 \tab 4 \tab - \tab 1 \tab 8 \tab - \tab 1 \tab 1 \tab - \tab 8 \tab - \tab 22 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 14 \tab - \tab 31 \tab 1 \tab 5 \tab - \tab - \tab 2 \tab 2 \tab 25 \tab 17 \tab 3 \tab - \tab 4 \tab 16 \tab 4 \tab 5 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 13 \tab 1 \tab 5 \tab 2 \tab 5 \tab 2 \tab - \tab 25 \tab 2 \tab 2 \tab 2 \tab 13 \tab 4 \tab 5 \tab 5 \tab 1 \tab - \tab 3 \tab 5 \tab 5 \tab - \tab 3 \tab 11 \tab 8 \tab 2 \tab 52 \tab 10 \tab 1 \tab - \tab - \tab 5 \tab 20 \tab - \tab 17 \tab - \tab 6 \tab 1 \tab - \tab - \tab 1 \tab - \tab 7 \tab 3 \tab 1 \tab 1 \tab 1 \tab 4 \tab 11 \tab 7 \tab 3 \tab 11 \tab 9 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 7 \tab 11 \tab 7 \tab 23 \tab 5 \tab 2 \tab 4 \tab - \tab 7 \tab 6 \tab 3 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab 18 \tab - \tab 8 \tab 1 \tab - \tab 4 \tab - \tab - \tab 52 \tab 14 \tab 11 \tab 12 \tab 27 \tab - \tab 8 \tab 7 \tab 1 \tab 9 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 4 \tab 7 \tab - \tab - \tab 24 \tab 3 \tab - \tab 3 \tab 2 \tab 4 \tab 6 \tab 1 \tab 12 \tab 4 \tab - \tab 2 \tab 38 \tab - \tab 18 \tab 12 \tab 8 \tab - \tab 8 \tab 3 \tab 5 \tab 2 \tab 2 \tab 12 \tab 29 \tab 5 \tab 5 \tab 2 \tab 4 \tab 1 \tab 7 \tab 4 \tab 1 \tab 7 \tab 4 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 5 \tab 3 \tab 6 \tab 2 \tab 4 \tab 8 \tab 1 \tab 3 \tab 4 \tab - \tab 7 \tab 80 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 5 \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 5 \tab 2 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 59 \tab 140 \tab 275 \tab 26 \tab 167 \tab 3 \tab 33 \tab 82 \tab 106 \tab 88 \tab 119 \tab 24 \tab 40 \tab 82 \tab 37 \tab 13 \tab 2 \tab 227 \tab 60 \tab 3 \tab 101 \tab 6 \tab 5 \tab 13 \tab 18 \tab 196 \tab 5 \tab 13 \tab 1 \tab 16 \tab 64 \tab 24 \tab 65 \tab 256 \tab 70 \tab 229 \tab 30 \tab 183 \tab 447 \tab 13 \tab 13 \tab 2 \tab 117 \tab 4 \tab 18 \tab 2 \tab 51 \tab 44 \tab 13 \tab 8 \tab 131 \tab 4 \tab 82 \tab 19 \tab 15 \tab 84 \tab 2 \tab 455 \tab 57 \tab 21 \tab 15 \tab 6 \tab 7 \tab 379 \tab 8 \tab 444 \tab 13 \tab 67 \tab 52 \tab 25 \tab 25 \tab 25 \tab 397 \tab 349 \tab 121 \tab 47 \tab 46 \tab 432 \tab 86 \tab 109 \tab 52 \tab 107 \tab 11 \tab 24 \tab 51 \tab 165 \tab 8 \tab 95 \tab 22 \tab 49 \tab 156 \tab 9 \tab 524 \tab 21 \tab 43 \tab 40 \tab 179 \tab 75 \tab 88 \tab 97 \tab 27 \tab 7 \tab 59 \tab 136 \tab 148 \tab 5 \tab 54 \tab 137 \tab 228 \tab 96 \tab 915 \tab 242 \tab 37 \tab 3 \tab 32 \tab 106 \tab 345 \tab 20 \tab 279 \tab 23 \tab 87 \tab 7 \tab 3 \tab 14 \tab 26 \tab 11 \tab 143 \tab 63 \tab 21 \tab 77 \tab 17 \tab 32 \tab 156 \tab 160 \tab 47 \tab 218 \tab 74 \tab 11 \tab 74 \tab 45 \tab 12 \tab 15 \tab 65 \tab 168 \tab 168 \tab 132 \tab 258 \tab 147 \tab 40 \tab 36 \tab 25 \tab 160 \tab 81 \tab 81 \tab 83 \tab 13 \tab 36 \tab 20 \tab 35 \tab 20 \tab 12 \tab 486 \tab 24 \tab 112 \tab 49 \tab 16 \tab 140 \tab 20 \tab 12 \tab 1066 \tab 116 \tab 88 \tab 142 \tab 424 \tab 29 \tab 93 \tab 119 \tab 47 \tab 159 \tab 27 \tab 54 \tab 41 \tab 57 \tab 8 \tab 40 \tab 6 \tab 73 \tab 75 \tab 58 \tab 27 \tab 33 \tab 367 \tab 33 \tab 15 \tab 44 \tab 33 \tab 60 \tab 71 \tab 25 \tab 150 \tab 75 \tab 10 \tab 56 \tab 682 \tab 38 \tab 228 \tab 124 \tab 97 \tab 44 \tab 46 \tab 23 \tab 46 \tab 72 \tab 50 \tab 344 \tab 524 \tab 42 \tab 47 \tab 36 \tab 86 \tab 29 \tab 85 \tab 43 \tab 20 \tab 64 \tab 57 \tab 36 \tab 11 \tab 43 \tab 11 \tab 10 \tab 18 \tab 13 \tab 4 \tab 29 \tab 10 \tab 69 \tab 10 \tab 23 \tab 93 \tab 27 \tab 10 \tab 52 \tab 13 \tab 91 \tab 48335 \tab 19 \tab 16 \tab 10 \tab 8 \tab 13 \tab 4 \tab 12 \tab 7 \tab 5 \tab 6 \tab 7 \tab 6 \tab 69 \tab 28 \tab 14 \tab 4 \tab 70 \tab 12 \tab 12 \tab 39 \tab 6 \tab 9 \tab 21 \tab 13 \tab 13 \tab 20 \tab 19 \tab 26 \tab 41 \tab 9 \tab 13 \tab 13 \tab 9 \tab 10 \tab 24 \tab 13 \tab 36 \tab 30 \tab 3 \tab 1 \tab 14 \tab 3 \tab 11 \tab 3 \tab 9 \tab 9 \tab 8 \tab 25 \tab 12 \tab 20 \tab 7 \tab 15 \tab 6 \tab 22 \tab 20 \tab 4 \tab 18 \tab 6 \tab 18 \tab 27 \tab 28 \tab 8 \tab 12 \tab 1 \tab 4 \tab 7 \tab 9 \tab 21 \tab 5 \tab 2 \tab 11 \tab 4 \tab 17 \tab 25 \tab 7 \tab 1 \tab 4 \tab 8 \tab 2 \tab 4 \tab 15 \tab 2 \tab 6 \tab 5 \tab 3 \tab 4 \tab 2 \tab 13 \tab 11 \tab 12 \tab 5 \tab 7 \tab 8 \tab 1 \tab 3 \tab 1 \tab 15 \tab 1 \tab 18 \tab 9 \tab 5 \tab 4 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 10 \tab 13 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
