% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_18.R
\name{dotthng}
\alias{dotthng}
\title{D.o.t. code - relation to things}
\source{
General Social Survey https://gss.norc.org
}
\description{
dotthng
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 0.21 \tab 0.26 \tab 0.31 \tab 0.85 \tab 1 \tab 1.1 \tab 1.15 \tab 1.16 \tab 1.25 \tab 1.26 \tab 1.28 \tab 1.36 \tab 1.38 \tab 1.4 \tab 1.41 \tab 1.42 \tab 1.43 \tab 1.45 \tab 1.46 \tab 1.49 \tab 1.51 \tab 1.55 \tab 1.57 \tab 1.59 \tab 1.62 \tab 1.64 \tab 1.65 \tab 1.68 \tab 1.74 \tab 1.87 \tab 1.88 \tab 1.9 \tab 1.91 \tab 1.96 \tab 1.97 \tab 1.98 \tab 2 \tab 2.09 \tab 2.15 \tab 2.32 \tab 2.33 \tab 2.39 \tab 2.43 \tab 2.48 \tab 2.52 \tab 2.53 \tab 2.6 \tab 2.61 \tab 2.65 \tab 2.67 \tab 2.72 \tab 2.76 \tab 2.82 \tab 2.84 \tab 2.85 \tab 2.87 \tab 2.91 \tab 2.93 \tab 3 \tab 3.01 \tab 3.02 \tab 3.03 \tab 3.05 \tab 3.07 \tab 3.13 \tab 3.16 \tab 3.26 \tab 3.29 \tab 3.35 \tab 3.45 \tab 3.47 \tab 3.52 \tab 3.65 \tab 3.68 \tab 3.71 \tab 3.72 \tab 3.74 \tab 3.78 \tab 3.81 \tab 3.83 \tab 4 \tab 4.03 \tab 4.06 \tab 4.07 \tab 4.08 \tab 4.14 \tab 4.18 \tab 4.2 \tab 4.22 \tab 4.25 \tab 4.32 \tab 4.34 \tab 4.35 \tab 4.36 \tab 4.39 \tab 4.42 \tab 4.5 \tab 4.53 \tab 4.59 \tab 4.61 \tab 4.64 \tab 4.67 \tab 4.7 \tab 4.72 \tab 4.74 \tab 4.78 \tab 4.81 \tab 4.93 \tab 4.96 \tab 5.07 \tab 5.09 \tab 5.1 \tab 5.13 \tab 5.14 \tab 5.15 \tab 5.2 \tab 5.21 \tab 5.3 \tab 5.35 \tab 5.37 \tab 5.39 \tab 5.42 \tab 5.53 \tab 5.58 \tab 5.63 \tab 5.76 \tab 5.84 \tab 5.87 \tab 5.88 \tab 5.91 \tab 5.93 \tab 6.08 \tab 6.09 \tab 6.1 \tab 6.13 \tab 6.2 \tab 6.34 \tab 6.38 \tab 6.48 \tab 6.51 \tab 6.63 \tab 6.66 \tab 6.68 \tab 6.72 \tab 6.82 \tab 6.83 \tab 6.84 \tab 6.85 \tab 6.9 \tab 6.93 \tab 6.96 \tab 7 \tab 7.06 \tab 7.07 \tab 7.12 \tab 7.13 \tab 7.16 \tab 7.18 \tab 7.19 \tab 7.22 \tab 7.4 \tab 7.41 \tab 7.44 \tab 7.47 \tab 7.49 \tab 7.5 \tab 7.51 \tab 7.52 \tab 7.55 \tab 7.57 \tab 7.58 \tab 7.61 \tab 7.62 \tab 7.64 \tab 7.65 \tab 7.66 \tab 7.67 \tab 7.69 \tab 7.7 \tab 7.71 \tab 7.73 \tab 7.74 \tab 7.76 \tab 7.77 \tab 7.78 \tab 7.79 \tab 7.8 \tab 7.81 \tab 7.82 \tab 7.84 \tab 7.85 \tab 7.86 \tab 7.87 \tab 7.88 \tab 7.89 \tab 7.91 \tab 7.93 \tab 7.94 \tab 7.95 \tab 7.96 \tab 7.97 \tab 7.98 \tab 7.99 \tab 8 \tab iap \tab 1.05 \tab 1.24 \tab 1.33 \tab 1.76 \tab 1.81 \tab 2.19 \tab 2.27 \tab 2.55 \tab 2.88 \tab 2.97 \tab 3.2 \tab 3.22 \tab 3.37 \tab 3.4 \tab 3.51 \tab 3.59 \tab 3.67 \tab 3.69 \tab 3.79 \tab 3.8 \tab 3.91 \tab 3.99 \tab 4.12 \tab 4.33 \tab 4.37 \tab 4.43 \tab 4.55 \tab 4.87 \tab 5.11 \tab 5.59 \tab 5.67 \tab 6 \tab 6.22 \tab 6.5 \tab 6.62 \tab 6.65 \tab 6.75 \tab 7.21 \tab 7.23 \tab 7.43 \tab 7.45 \tab 7.53 \tab 7.72 \tab 7.92 \tab 0.56 \tab 0.95 \tab 1.39 \tab 2.21 \tab 2.74 \tab 2.8 \tab 3.08 \tab 3.21 \tab 4.01 \tab 4.05 \tab 4.28 \tab 4.79 \tab 5.22 \tab 5.82 \tab 5.95 \tab 6.43 \tab 6.46 \tab 6.58 \tab 7.48 \tab 7.68 \tab 7.83 \tab 0 \tab 0.55 \tab 1.71 \tab 2.22 \tab 3.95 \tab 4.3 \tab 4.38 \tab 5.45 \tab 5.47 \tab 5.89 \tab 6.17 \tab 7.02 \tab 7.28 \tab 0.79 \tab 1.32 \tab 1.5 \tab 2.71 \tab 6.15 \tab 6.49 \tab 6.69 \tab 1.11 \tab 3.41 \tab 4.75 \tab 4.89 \tab 5.32 \tab 5.5 \tab 5.8 \tab 5.86 \tab 6.14 \tab 1.61 \tab 4.71 \tab 6.18 \tab 6.23 \tab 7.63 \tab 1.84 \tab 3.54 \tab 4.15 \tab 6.3 \tab 6.55 \tab 2.16 \tab 3.93 \tab 6.42 \tab 7.25 \tab 7.26 \tab 4.62 \tab 1.79 \tab 2.01 \tab 2.57 \tab 3.48 \tab 5.24 \tab Total \cr
   1972 \tab 10 \tab 1 \tab 5 \tab 1 \tab 3 \tab 10 \tab 6 \tab 7 \tab 18 \tab 13 \tab 1 \tab 3 \tab 2 \tab 1 \tab 7 \tab 4 \tab 3 \tab 3 \tab 10 \tab 1 \tab 39 \tab 2 \tab 2 \tab 1 \tab 3 \tab 20 \tab 2 \tab 5 \tab 3 \tab 12 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 7 \tab 3 \tab 1 \tab 1 \tab 17 \tab 1 \tab 7 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 12 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab 5 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 6 \tab 4 \tab 48 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 11 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 7 \tab 1 \tab 2 \tab 6 \tab 7 \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 4 \tab 1 \tab 5 \tab 5 \tab 1 \tab 15 \tab 3 \tab 3 \tab 1 \tab 5 \tab 2 \tab 14 \tab 4 \tab 18 \tab 2 \tab 15 \tab 3 \tab 3 \tab 1 \tab 1 \tab 8 \tab 1 \tab 5 \tab 1 \tab 11 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 5 \tab 5 \tab 1 \tab 8 \tab 1 \tab 2 \tab 1 \tab 5 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2 \tab 2 \tab 15 \tab 1 \tab 7 \tab 2 \tab 3 \tab 14 \tab 8 \tab 2 \tab 1 \tab 1 \tab 5 \tab 12 \tab 1 \tab 1 \tab 3 \tab 6 \tab 8 \tab 1 \tab 2 \tab 21 \tab 1 \tab 2 \tab 3 \tab 34 \tab 6 \tab 1 \tab 2 \tab 2 \tab 6 \tab 12 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 6 \tab 3 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 14 \tab 5 \tab 31 \tab 4 \tab 7 \tab 1 \tab 18 \tab 2 \tab 9 \tab 29 \tab 37 \tab 6 \tab 5 \tab 16 \tab 20 \tab 54 \tab 13 \tab 5 \tab 34 \tab 47 \tab 72 \tab 40 \tab 151 \tab 183 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 13 \tab 1 \tab 2 \tab - \tab 7 \tab 6 \tab 3 \tab 4 \tab 13 \tab 20 \tab 3 \tab 4 \tab 1 \tab 1 \tab 8 \tab 5 \tab 3 \tab 3 \tab 3 \tab - \tab 24 \tab 2 \tab 2 \tab - \tab 2 \tab 15 \tab 7 \tab 4 \tab 4 \tab 13 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 17 \tab 3 \tab 4 \tab 3 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab 6 \tab 1 \tab - \tab - \tab 3 \tab - \tab 2 \tab 3 \tab 6 \tab 2 \tab 3 \tab 3 \tab - \tab 3 \tab 24 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 6 \tab - \tab 2 \tab 4 \tab 1 \tab 5 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 5 \tab - \tab 16 \tab 1 \tab 3 \tab - \tab 9 \tab 1 \tab 8 \tab 7 \tab 22 \tab - \tab 7 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab 12 \tab - \tab 1 \tab 3 \tab 3 \tab 4 \tab 5 \tab 5 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 4 \tab 1 \tab 14 \tab 4 \tab 5 \tab 1 \tab - \tab 6 \tab 9 \tab - \tab - \tab 1 \tab 4 \tab 9 \tab 4 \tab 2 \tab 2 \tab 7 \tab 9 \tab 6 \tab 3 \tab 33 \tab - \tab 1 \tab 7 \tab 33 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 8 \tab 2 \tab - \tab 2 \tab 4 \tab 2 \tab 8 \tab 2 \tab 1 \tab 11 \tab 1 \tab 2 \tab 2 \tab 3 \tab 10 \tab 6 \tab 17 \tab 1 \tab 8 \tab - \tab 14 \tab - \tab 10 \tab 25 \tab 26 \tab 8 \tab 4 \tab 11 \tab 33 \tab 37 \tab 13 \tab 2 \tab 43 \tab 42 \tab 85 \tab 39 \tab 125 \tab 183 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 2 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 2 \tab 1 \tab 1 \tab 1 \tab 5 \tab 9 \tab 7 \tab 2 \tab 19 \tab 21 \tab 1 \tab - \tab 3 \tab 1 \tab 13 \tab 2 \tab 4 \tab 5 \tab 3 \tab - \tab 20 \tab 1 \tab 3 \tab 1 \tab 3 \tab 18 \tab 3 \tab 6 \tab - \tab 10 \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 4 \tab 6 \tab - \tab 1 \tab 17 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 11 \tab 3 \tab - \tab - \tab 3 \tab 5 \tab 5 \tab 2 \tab 3 \tab 1 \tab 4 \tab 8 \tab 3 \tab 2 \tab 15 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab - \tab 2 \tab - \tab 4 \tab 5 \tab 3 \tab - \tab 7 \tab 2 \tab - \tab 5 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 2 \tab 2 \tab 20 \tab 3 \tab 1 \tab - \tab 4 \tab 4 \tab 9 \tab 1 \tab 16 \tab - \tab 15 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 3 \tab 10 \tab - \tab 1 \tab 1 \tab 5 \tab 2 \tab 8 \tab 5 \tab - \tab 10 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 7 \tab - \tab 8 \tab 1 \tab 1 \tab 1 \tab - \tab 13 \tab 4 \tab - \tab - \tab 2 \tab 3 \tab 8 \tab 6 \tab 1 \tab 4 \tab 5 \tab 12 \tab 7 \tab - \tab 22 \tab - \tab 2 \tab 5 \tab 20 \tab 1 \tab - \tab 4 \tab 1 \tab 3 \tab 7 \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab 7 \tab 5 \tab 2 \tab 4 \tab 4 \tab - \tab - \tab 2 \tab 9 \tab 4 \tab 21 \tab 3 \tab 9 \tab 2 \tab 11 \tab 1 \tab 3 \tab 48 \tab 3 \tab 6 \tab 10 \tab 14 \tab 12 \tab 45 \tab 14 \tab 10 \tab 43 \tab 54 \tab 78 \tab 48 \tab 172 \tab 151 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab 6 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 7 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 11 \tab 1 \tab 1 \tab - \tab 4 \tab 9 \tab 2 \tab 1 \tab 14 \tab 22 \tab 1 \tab 1 \tab 1 \tab 1 \tab 14 \tab 3 \tab 6 \tab 3 \tab 10 \tab - \tab 22 \tab 3 \tab 5 \tab - \tab 2 \tab 15 \tab 2 \tab 7 \tab 2 \tab 12 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 3 \tab 7 \tab - \tab 2 \tab 38 \tab 1 \tab 3 \tab 3 \tab 2 \tab 5 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 7 \tab 2 \tab - \tab - \tab - \tab - \tab 4 \tab 4 \tab 3 \tab 1 \tab 3 \tab 5 \tab 4 \tab 4 \tab 22 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 7 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 2 \tab 2 \tab 6 \tab - \tab 1 \tab 5 \tab 8 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 2 \tab 1 \tab 20 \tab 2 \tab 3 \tab - \tab 5 \tab 1 \tab 8 \tab 1 \tab 18 \tab - \tab 7 \tab - \tab 1 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 8 \tab - \tab - \tab 2 \tab 4 \tab 2 \tab 6 \tab 1 \tab - \tab 9 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 12 \tab 2 \tab 3 \tab 2 \tab 2 \tab 5 \tab 6 \tab - \tab - \tab - \tab 3 \tab 8 \tab 2 \tab 1 \tab 1 \tab 7 \tab 15 \tab 10 \tab 1 \tab 26 \tab 1 \tab 2 \tab 8 \tab 26 \tab 1 \tab 2 \tab 3 \tab - \tab 4 \tab 3 \tab 1 \tab 3 \tab 1 \tab 5 \tab - \tab 4 \tab 2 \tab 3 \tab 2 \tab 3 \tab 5 \tab - \tab - \tab 9 \tab 3 \tab 26 \tab 4 \tab 6 \tab - \tab 7 \tab - \tab 4 \tab 15 \tab 54 \tab 13 \tab 4 \tab 14 \tab 12 \tab 47 \tab 7 \tab 14 \tab 43 \tab 56 \tab 81 \tab 41 \tab 137 \tab 155 \tab - \tab 4 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 6 \tab - \tab - \tab 5 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 7 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 4 \tab - \tab 3 \tab - \tab 4 \tab 11 \tab 5 \tab 2 \tab 9 \tab 7 \tab 2 \tab 2 \tab 1 \tab 1 \tab 14 \tab 2 \tab 3 \tab 4 \tab 8 \tab 1 \tab 37 \tab 2 \tab 2 \tab - \tab 1 \tab 23 \tab 5 \tab 2 \tab 7 \tab 6 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 30 \tab 2 \tab 4 \tab 4 \tab 2 \tab 5 \tab - \tab - \tab 1 \tab 1 \tab - \tab 14 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab 8 \tab - \tab 2 \tab 4 \tab 5 \tab 3 \tab 18 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 13 \tab 1 \tab 3 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab 5 \tab - \tab 2 \tab 2 \tab 4 \tab 4 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab 25 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 12 \tab 6 \tab 13 \tab - \tab 14 \tab 3 \tab - \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 8 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 14 \tab 7 \tab - \tab 9 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 4 \tab 2 \tab 10 \tab 3 \tab 5 \tab 3 \tab - \tab 11 \tab 8 \tab - \tab 2 \tab 2 \tab 3 \tab 8 \tab 3 \tab - \tab - \tab 11 \tab 11 \tab 3 \tab 3 \tab 18 \tab - \tab 1 \tab 1 \tab 32 \tab 2 \tab 1 \tab 4 \tab 3 \tab 1 \tab 8 \tab - \tab 2 \tab 3 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 7 \tab 1 \tab 4 \tab 2 \tab 1 \tab 12 \tab 8 \tab 21 \tab - \tab 3 \tab 2 \tab 11 \tab - \tab 3 \tab 20 \tab 63 \tab 9 \tab 3 \tab 11 \tab 16 \tab 39 \tab 8 \tab 5 \tab 51 \tab 56 \tab 77 \tab 44 \tab 128 \tab 158 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 6 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 4 \tab 1 \tab 1 \tab - \tab 7 \tab 10 \tab 3 \tab 2 \tab 12 \tab 22 \tab 1 \tab 3 \tab - \tab 2 \tab 20 \tab 2 \tab 3 \tab 1 \tab 13 \tab 1 \tab 23 \tab 1 \tab 4 \tab - \tab 2 \tab 16 \tab 3 \tab 5 \tab 7 \tab 12 \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab 3 \tab 2 \tab - \tab - \tab 25 \tab 1 \tab 1 \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 10 \tab 2 \tab 3 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 4 \tab 3 \tab 2 \tab 3 \tab 10 \tab 5 \tab 21 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 14 \tab - \tab 4 \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab 6 \tab 4 \tab 3 \tab 8 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab - \tab 27 \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab 11 \tab - \tab 16 \tab - \tab 14 \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab - \tab 2 \tab 17 \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab 3 \tab 3 \tab - \tab 16 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 25 \tab 3 \tab 3 \tab - \tab 1 \tab 15 \tab 16 \tab 1 \tab - \tab 2 \tab 1 \tab 5 \tab 1 \tab - \tab 1 \tab 5 \tab 9 \tab 6 \tab - \tab 20 \tab 1 \tab 2 \tab 1 \tab 28 \tab 6 \tab 3 \tab 5 \tab - \tab 2 \tab 8 \tab - \tab - \tab 3 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab 16 \tab 4 \tab 22 \tab 3 \tab 6 \tab 1 \tab 15 \tab 1 \tab 2 \tab 7 \tab 58 \tab 4 \tab 7 \tab 13 \tab 16 \tab 41 \tab 8 \tab 8 \tab 31 \tab 61 \tab 83 \tab 35 \tab 156 \tab 133 \tab 1 \tab 3 \tab 2 \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab - \tab 6 \tab 2 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 4 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 6 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 5 \tab 3 \tab 6 \tab 1 \tab 5 \tab 11 \tab 5 \tab 1 \tab 16 \tab 17 \tab 2 \tab 3 \tab - \tab 4 \tab 15 \tab 4 \tab - \tab 2 \tab 8 \tab - \tab 15 \tab 2 \tab 6 \tab 2 \tab - \tab 18 \tab 3 \tab 4 \tab 3 \tab 16 \tab - \tab 5 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 31 \tab 5 \tab 4 \tab 4 \tab - \tab 4 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 10 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 5 \tab 1 \tab 3 \tab - \tab - \tab 8 \tab 4 \tab 5 \tab 24 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab 8 \tab 1 \tab 2 \tab - \tab 5 \tab 2 \tab 7 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 5 \tab - \tab 1 \tab 3 \tab 3 \tab 19 \tab - \tab 3 \tab - \tab 4 \tab 3 \tab 10 \tab 6 \tab 12 \tab 1 \tab 19 \tab - \tab 1 \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab 5 \tab - \tab 2 \tab 4 \tab 10 \tab 2 \tab 4 \tab 3 \tab - \tab 11 \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 12 \tab - \tab 3 \tab 4 \tab 1 \tab 14 \tab 4 \tab - \tab - \tab 4 \tab 2 \tab 10 \tab 3 \tab 2 \tab 2 \tab 8 \tab 10 \tab 5 \tab - \tab 22 \tab 1 \tab 2 \tab 3 \tab 24 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab - \tab 3 \tab 2 \tab 2 \tab 5 \tab 2 \tab 5 \tab 4 \tab - \tab 5 \tab 1 \tab 1 \tab 13 \tab 2 \tab 22 \tab 6 \tab 3 \tab - \tab 16 \tab - \tab 5 \tab 15 \tab 63 \tab 5 \tab 3 \tab 2 \tab 15 \tab 48 \tab 5 \tab 6 \tab 58 \tab 44 \tab 88 \tab 45 \tab 157 \tab 149 \tab 2 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 6 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 6 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 5 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 13 \tab - \tab 4 \tab - \tab 6 \tab 7 \tab 7 \tab 2 \tab 11 \tab 12 \tab 2 \tab 1 \tab 3 \tab 1 \tab 9 \tab - \tab 3 \tab 3 \tab 9 \tab - \tab 24 \tab 1 \tab - \tab - \tab - \tab 19 \tab 3 \tab 2 \tab 9 \tab 8 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab - \tab 1 \tab 23 \tab 4 \tab - \tab 5 \tab 1 \tab 9 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 5 \tab - \tab 2 \tab - \tab 3 \tab 3 \tab 6 \tab 2 \tab 5 \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 18 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 12 \tab 1 \tab 3 \tab 3 \tab 6 \tab 3 \tab 3 \tab 2 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 2 \tab 17 \tab - \tab - \tab 2 \tab 3 \tab 7 \tab 8 \tab 4 \tab 21 \tab - \tab 8 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 7 \tab 1 \tab - \tab 2 \tab 5 \tab 1 \tab 5 \tab 3 \tab 1 \tab 9 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 1 \tab 3 \tab 1 \tab 8 \tab 1 \tab 2 \tab 5 \tab - \tab 6 \tab 9 \tab 1 \tab - \tab 3 \tab 3 \tab 11 \tab 2 \tab - \tab - \tab 2 \tab 9 \tab 6 \tab 1 \tab 11 \tab 1 \tab 1 \tab 8 \tab 25 \tab 3 \tab 3 \tab 5 \tab 1 \tab 4 \tab 6 \tab - \tab - \tab 2 \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab 11 \tab 2 \tab 12 \tab - \tab - \tab 21 \tab 4 \tab 13 \tab 1 \tab 7 \tab - \tab 13 \tab - \tab 11 \tab 19 \tab 44 \tab 9 \tab 6 \tab 7 \tab 17 \tab 52 \tab 9 \tab 13 \tab 45 \tab 37 \tab 83 \tab 42 \tab 155 \tab 133 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 13 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 4 \tab 1 \tab 3 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 12 \tab 3 \tab 5 \tab 19 \tab 14 \tab 10 \tab - \tab 3 \tab 3 \tab 16 \tab 2 \tab 5 \tab 6 \tab 8 \tab - \tab 21 \tab 2 \tab 7 \tab 3 \tab 1 \tab 17 \tab 8 \tab 1 \tab 10 \tab 19 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 22 \tab 6 \tab 2 \tab 2 \tab - \tab 6 \tab 1 \tab - \tab 4 \tab - \tab - \tab 10 \tab 4 \tab 3 \tab 1 \tab 1 \tab 13 \tab 3 \tab - \tab 2 \tab - \tab 3 \tab 4 \tab 7 \tab 4 \tab 26 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab 9 \tab 1 \tab 4 \tab - \tab 1 \tab 5 \tab 5 \tab 2 \tab 10 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab - \tab 5 \tab - \tab - \tab 12 \tab 1 \tab 3 \tab 3 \tab - \tab 22 \tab - \tab 5 \tab 2 \tab 8 \tab 1 \tab 13 \tab 2 \tab 13 \tab - \tab 20 \tab 3 \tab - \tab - \tab 1 \tab 4 \tab 2 \tab 4 \tab 4 \tab 19 \tab - \tab 3 \tab 2 \tab 6 \tab 3 \tab 9 \tab 7 \tab - \tab 7 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 4 \tab 2 \tab 1 \tab 19 \tab 3 \tab 7 \tab 3 \tab 1 \tab 17 \tab 27 \tab - \tab - \tab 1 \tab - \tab 40 \tab 3 \tab 2 \tab 5 \tab 3 \tab 17 \tab 7 \tab - \tab 11 \tab 1 \tab - \tab 3 \tab 45 \tab 3 \tab 2 \tab 16 \tab - \tab 1 \tab 14 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 5 \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab 13 \tab 4 \tab 51 \tab 7 \tab 10 \tab - \tab 5 \tab 1 \tab 14 \tab 15 \tab 59 \tab 2 \tab 4 \tab 18 \tab 26 \tab 53 \tab 23 \tab 4 \tab 40 \tab 88 \tab 80 \tab 54 \tab 188 \tab 148 \tab 3 \tab - \tab 3 \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 7 \tab 2 \tab 1 \tab 7 \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 5 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab - \tab 3 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 8 \tab 1 \tab 1 \tab 4 \tab 12 \tab 5 \tab 4 \tab 2 \tab 22 \tab 14 \tab 3 \tab 4 \tab 3 \tab 1 \tab 10 \tab 1 \tab 5 \tab 1 \tab 6 \tab - \tab 33 \tab 2 \tab 4 \tab - \tab 2 \tab 29 \tab 4 \tab 1 \tab 9 \tab 9 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 23 \tab 2 \tab - \tab 2 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab 11 \tab 3 \tab 1 \tab 6 \tab - \tab 4 \tab 3 \tab 2 \tab 1 \tab 19 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 10 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab 2 \tab 5 \tab 4 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 25 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 8 \tab 1 \tab 15 \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab 10 \tab - \tab - \tab - \tab 8 \tab 7 \tab 6 \tab 6 \tab - \tab 7 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 2 \tab 10 \tab 2 \tab 6 \tab 8 \tab 1 \tab 9 \tab 10 \tab 1 \tab - \tab 2 \tab 2 \tab 4 \tab 2 \tab - \tab 3 \tab 7 \tab 2 \tab 12 \tab 1 \tab 17 \tab - \tab 1 \tab 10 \tab 36 \tab 4 \tab - \tab 7 \tab 1 \tab 7 \tab 4 \tab - \tab 2 \tab 3 \tab 2 \tab 4 \tab 4 \tab 5 \tab 2 \tab 2 \tab 4 \tab 6 \tab - \tab 1 \tab 10 \tab 10 \tab 6 \tab 4 \tab 6 \tab - \tab 25 \tab 6 \tab 4 \tab 13 \tab 58 \tab 3 \tab 3 \tab 13 \tab 20 \tab 39 \tab 8 \tab 11 \tab 59 \tab 69 \tab 123 \tab 52 \tab 164 \tab 118 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab 6 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 9 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 8 \tab - \tab 1 \tab 3 \tab 4 \tab 4 \tab 2 \tab 1 \tab 11 \tab 17 \tab 2 \tab 4 \tab 2 \tab 1 \tab 13 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 21 \tab - \tab 6 \tab - \tab - \tab 24 \tab 5 \tab 1 \tab 13 \tab 13 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 16 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 7 \tab 2 \tab - \tab - \tab 1 \tab 11 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 2 \tab 6 \tab 22 \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 10 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 16 \tab - \tab 6 \tab 2 \tab 2 \tab 1 \tab 14 \tab 2 \tab 13 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 2 \tab 6 \tab - \tab 1 \tab 8 \tab 6 \tab - \tab - \tab 1 \tab 2 \tab 16 \tab 5 \tab 4 \tab - \tab 5 \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 13 \tab 3 \tab - \tab 1 \tab 1 \tab 9 \tab 2 \tab - \tab 1 \tab 5 \tab 3 \tab 16 \tab 5 \tab - \tab 1 \tab 9 \tab 17 \tab 13 \tab 2 \tab 21 \tab 3 \tab - \tab 6 \tab 33 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 3 \tab - \tab 28 \tab 1 \tab 7 \tab 3 \tab 2 \tab 7 \tab 1 \tab - \tab 19 \tab 2 \tab 7 \tab 3 \tab 8 \tab 1 \tab 35 \tab 3 \tab 3 \tab 11 \tab 56 \tab 9 \tab 1 \tab 11 \tab 21 \tab 46 \tab 5 \tab 5 \tab 35 \tab 51 \tab 77 \tab 47 \tab 162 \tab 113 \tab - \tab 6 \tab 2 \tab 1 \tab 4 \tab 6 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 3 \tab 5 \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 6 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 6 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 3 \tab 1 \tab 4 \tab - \tab 8 \tab 13 \tab 2 \tab 2 \tab 23 \tab 10 \tab 7 \tab 6 \tab 2 \tab 2 \tab 10 \tab 2 \tab 6 \tab 4 \tab 6 \tab - \tab 24 \tab 2 \tab 5 \tab 1 \tab 2 \tab 22 \tab 1 \tab 1 \tab 6 \tab 14 \tab - \tab 1 \tab - \tab 5 \tab 4 \tab - \tab 7 \tab - \tab - \tab 18 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab 6 \tab 3 \tab 1 \tab - \tab 1 \tab 17 \tab 3 \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 6 \tab 3 \tab 29 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 8 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 13 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 13 \tab 3 \tab 13 \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab 14 \tab 2 \tab - \tab 1 \tab 10 \tab 1 \tab 6 \tab 5 \tab - \tab 2 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 7 \tab - \tab - \tab 2 \tab 10 \tab 1 \tab 2 \tab 4 \tab 1 \tab 7 \tab 13 \tab 4 \tab - \tab 6 \tab 3 \tab 23 \tab 2 \tab 2 \tab 2 \tab 5 \tab 14 \tab 6 \tab 4 \tab 10 \tab 1 \tab 2 \tab 6 \tab 21 \tab 3 \tab 3 \tab 1 \tab 3 \tab 2 \tab 9 \tab - \tab 2 \tab 4 \tab - \tab 6 \tab 4 \tab 2 \tab 9 \tab 4 \tab 1 \tab 4 \tab 1 \tab 3 \tab 24 \tab 10 \tab 16 \tab 6 \tab 9 \tab 4 \tab 12 \tab 10 \tab 2 \tab 15 \tab 46 \tab 15 \tab 8 \tab 17 \tab 15 \tab 60 \tab 12 \tab 6 \tab 48 \tab 52 \tab 78 \tab 51 \tab 166 \tab 85 \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 6 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 10 \tab - \tab 4 \tab - \tab 5 \tab 4 \tab 3 \tab 1 \tab 9 \tab 16 \tab 2 \tab 3 \tab 1 \tab 1 \tab 9 \tab 2 \tab 2 \tab 5 \tab 7 \tab - \tab 28 \tab 3 \tab 3 \tab - \tab 1 \tab 18 \tab 11 \tab 1 \tab 6 \tab 16 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab - \tab 1 \tab 11 \tab 5 \tab 3 \tab 4 \tab 1 \tab 2 \tab 5 \tab 1 \tab 3 \tab 2 \tab - \tab 9 \tab 1 \tab 2 \tab - \tab 1 \tab 9 \tab 5 \tab - \tab 5 \tab - \tab 2 \tab 4 \tab 1 \tab 4 \tab 22 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 10 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 16 \tab - \tab 2 \tab 1 \tab 1 \tab 13 \tab 6 \tab 1 \tab 11 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 8 \tab 9 \tab 8 \tab 7 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 6 \tab 1 \tab 5 \tab 1 \tab 8 \tab 5 \tab - \tab 4 \tab - \tab 4 \tab 9 \tab - \tab 1 \tab 6 \tab 4 \tab 5 \tab - \tab 4 \tab 1 \tab 6 \tab 9 \tab 8 \tab 3 \tab 18 \tab 1 \tab 2 \tab 12 \tab 41 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 7 \tab 1 \tab 2 \tab 3 \tab 4 \tab 1 \tab 2 \tab 1 \tab 2 \tab 6 \tab 1 \tab 4 \tab - \tab 1 \tab 19 \tab 12 \tab 4 \tab 3 \tab 7 \tab 2 \tab 25 \tab 1 \tab 3 \tab 14 \tab 65 \tab 10 \tab 6 \tab 9 \tab 27 \tab 51 \tab 6 \tab 8 \tab 25 \tab 56 \tab 69 \tab 55 \tab 162 \tab 109 \tab - \tab 1 \tab 2 \tab 1 \tab 6 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 4 \tab 2 \tab - \tab 1 \tab 7 \tab 9 \tab 7 \tab 2 \tab 17 \tab 20 \tab - \tab 3 \tab 2 \tab 2 \tab 9 \tab 3 \tab 2 \tab 4 \tab 7 \tab 2 \tab 31 \tab 5 \tab 9 \tab 1 \tab 1 \tab 25 \tab 9 \tab 1 \tab 12 \tab 19 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 6 \tab 1 \tab - \tab 23 \tab 4 \tab 2 \tab 6 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 4 \tab - \tab 7 \tab - \tab 1 \tab - \tab - \tab 18 \tab 10 \tab - \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 19 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 10 \tab - \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 20 \tab 1 \tab 1 \tab 1 \tab 6 \tab 1 \tab 11 \tab 3 \tab 22 \tab - \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 3 \tab 7 \tab 24 \tab - \tab - \tab 1 \tab 6 \tab 10 \tab 7 \tab 6 \tab - \tab 11 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 5 \tab 2 \tab 10 \tab - \tab 5 \tab 2 \tab 1 \tab 6 \tab 28 \tab 1 \tab - \tab 5 \tab 10 \tab 9 \tab 3 \tab 5 \tab 3 \tab 3 \tab 16 \tab 16 \tab 1 \tab 22 \tab - \tab - \tab 22 \tab 61 \tab 13 \tab 4 \tab 4 \tab 5 \tab 1 \tab 8 \tab 1 \tab 4 \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 5 \tab - \tab 2 \tab 19 \tab 20 \tab 14 \tab 9 \tab 15 \tab 2 \tab 30 \tab 3 \tab 7 \tab 11 \tab 59 \tab 13 \tab 3 \tab 19 \tab 34 \tab 37 \tab 11 \tab 5 \tab 27 \tab 88 \tab 109 \tab 45 \tab 169 \tab 113 \tab 1 \tab 3 \tab 1 \tab 1 \tab 5 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 9 \tab 2 \tab 4 \tab - \tab - \tab 5 \tab 4 \tab - \tab 1 \tab 4 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab - \tab 3 \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1819 \cr
   1988 \tab 4 \tab 1 \tab 1 \tab 3 \tab 12 \tab 4 \tab 11 \tab 1 \tab 10 \tab 17 \tab 3 \tab 4 \tab 3 \tab - \tab 21 \tab 3 \tab 3 \tab 2 \tab 8 \tab - \tab 19 \tab 2 \tab 5 \tab - \tab - \tab 28 \tab 3 \tab - \tab 3 \tab 8 \tab 2 \tab 4 \tab 2 \tab 1 \tab 2 \tab 2 \tab 7 \tab - \tab - \tab 27 \tab 3 \tab 4 \tab - \tab 2 \tab 3 \tab 3 \tab - \tab - \tab 8 \tab 1 \tab 8 \tab 2 \tab 1 \tab - \tab - \tab 12 \tab 3 \tab 1 \tab - \tab - \tab - \tab 10 \tab 4 \tab 1 \tab 20 \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 5 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 5 \tab 2 \tab 2 \tab 21 \tab - \tab 1 \tab 1 \tab 5 \tab 2 \tab 2 \tab 4 \tab 6 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 14 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab 4 \tab 4 \tab - \tab 6 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 6 \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab 16 \tab - \tab - \tab 5 \tab 3 \tab 22 \tab 1 \tab 1 \tab 3 \tab 6 \tab 17 \tab 11 \tab - \tab 11 \tab - \tab 2 \tab 6 \tab 33 \tab 3 \tab 4 \tab 4 \tab 4 \tab 1 \tab 7 \tab - \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab 3 \tab 2 \tab 5 \tab - \tab 1 \tab 14 \tab 6 \tab 6 \tab 7 \tab 16 \tab 3 \tab 19 \tab 5 \tab 2 \tab 9 \tab 69 \tab 16 \tab 7 \tab 12 \tab 21 \tab 46 \tab 4 \tab 6 \tab 28 \tab 60 \tab 80 \tab 45 \tab 164 \tab 93 \tab 3 \tab 2 \tab 2 \tab 4 \tab 3 \tab 6 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 8 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 6 \tab - \tab 2 \tab 2 \tab 6 \tab 12 \tab 3 \tab 5 \tab 21 \tab 15 \tab 2 \tab 7 \tab 5 \tab 2 \tab 12 \tab 3 \tab 2 \tab 4 \tab 9 \tab 2 \tab 18 \tab 4 \tab 8 \tab 2 \tab 5 \tab 19 \tab 2 \tab 7 \tab 5 \tab 6 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 5 \tab 5 \tab 1 \tab - \tab 9 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 5 \tab 2 \tab - \tab - \tab 1 \tab 6 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 5 \tab 1 \tab 3 \tab 22 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 9 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 13 \tab 2 \tab 3 \tab 1 \tab 5 \tab 1 \tab 4 \tab 2 \tab 14 \tab 1 \tab 6 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab 1 \tab 11 \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab 4 \tab 9 \tab - \tab 10 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 7 \tab 1 \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 5 \tab 1 \tab 7 \tab 12 \tab - \tab - \tab 3 \tab 5 \tab 16 \tab 2 \tab 4 \tab 2 \tab 8 \tab 15 \tab 10 \tab 3 \tab 11 \tab 2 \tab - \tab 6 \tab 38 \tab 3 \tab 5 \tab 1 \tab 3 \tab 6 \tab 11 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 2 \tab 5 \tab 8 \tab - \tab 4 \tab 2 \tab 2 \tab 22 \tab 8 \tab 10 \tab 3 \tab 13 \tab 1 \tab 15 \tab 7 \tab 7 \tab 10 \tab 63 \tab 11 \tab 4 \tab 11 \tab 16 \tab 48 \tab 11 \tab 5 \tab 30 \tab 73 \tab 85 \tab 61 \tab 185 \tab 106 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 5 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1537 \cr
   1990 \tab 6 \tab - \tab 1 \tab - \tab 5 \tab 7 \tab 1 \tab 1 \tab 9 \tab 16 \tab 3 \tab 2 \tab 4 \tab 5 \tab 8 \tab 4 \tab 3 \tab 7 \tab 3 \tab - \tab 16 \tab 3 \tab 4 \tab 2 \tab 1 \tab 19 \tab 10 \tab 4 \tab 5 \tab 11 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 7 \tab - \tab - \tab 13 \tab - \tab 1 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab 2 \tab - \tab - \tab 6 \tab - \tab - \tab 3 \tab 1 \tab - \tab 6 \tab 3 \tab 2 \tab 33 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 3 \tab 15 \tab 3 \tab 2 \tab 1 \tab 6 \tab 1 \tab 7 \tab 2 \tab 13 \tab - \tab 13 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 2 \tab 12 \tab 1 \tab - \tab 2 \tab 4 \tab 3 \tab 6 \tab - \tab - \tab 4 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 11 \tab 1 \tab 2 \tab 1 \tab - \tab 8 \tab 8 \tab 1 \tab - \tab 4 \tab 6 \tab 9 \tab 3 \tab 1 \tab - \tab 10 \tab 6 \tab 6 \tab - \tab 18 \tab - \tab - \tab 8 \tab 29 \tab 2 \tab 4 \tab 3 \tab 3 \tab 1 \tab 5 \tab - \tab 8 \tab 8 \tab 5 \tab 1 \tab 5 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 29 \tab 5 \tab 7 \tab 1 \tab 10 \tab 4 \tab 17 \tab - \tab 5 \tab 6 \tab 52 \tab 11 \tab 6 \tab 12 \tab 24 \tab 36 \tab 13 \tab 6 \tab 33 \tab 59 \tab 82 \tab 46 \tab 161 \tab 80 \tab - \tab 2 \tab 3 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab 5 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 112 \tab 14 \tab 39 \tab 17 \tab 106 \tab 143 \tab 74 \tab 41 \tab 253 \tab 273 \tab 45 \tab 50 \tab 36 \tab 29 \tab 208 \tab 43 \tab 54 \tab 58 \tab 121 \tab 7 \tab 415 \tab 37 \tab 75 \tab 13 \tab 26 \tab 345 \tab 81 \tab 52 \tab 104 \tab 204 \tab 7 \tab 30 \tab 10 \tab 18 \tab 37 \tab 35 \tab 67 \tab 4 \tab 12 \tab 360 \tab 41 \tab 39 \tab 53 \tab 15 \tab 55 \tab 23 \tab 20 \tab 20 \tab 37 \tab 5 \tab 136 \tab 31 \tab 19 \tab 2 \tab 20 \tab 121 \tab 61 \tab 16 \tab 64 \tab 11 \tab 29 \tab 73 \tab 61 \tab 51 \tab 402 \tab 25 \tab 8 \tab 14 \tab 13 \tab 21 \tab 7 \tab 18 \tab 21 \tab 15 \tab 148 \tab 10 \tab 28 \tab 15 \tab 38 \tab 50 \tab 37 \tab 24 \tab 72 \tab 15 \tab 19 \tab 44 \tab 55 \tab 30 \tab 16 \tab 8 \tab 29 \tab 20 \tab 4 \tab 48 \tab 9 \tab 33 \tab 36 \tab 20 \tab 320 \tab 17 \tab 40 \tab 19 \tab 75 \tab 41 \tab 158 \tab 49 \tab 256 \tab 4 \tab 160 \tab 13 \tab 12 \tab 1 \tab 13 \tab 56 \tab 17 \tab 32 \tab 40 \tab 190 \tab 9 \tab 9 \tab 28 \tab 86 \tab 70 \tab 105 \tab 80 \tab 2 \tab 131 \tab 29 \tab 13 \tab 2 \tab 8 \tab 2 \tab 16 \tab 7 \tab 35 \tab 30 \tab 45 \tab 17 \tab 195 \tab 33 \tab 53 \tab 49 \tab 13 \tab 154 \tab 189 \tab 11 \tab 5 \tab 52 \tab 60 \tab 215 \tab 43 \tab 26 \tab 33 \tab 108 \tab 196 \tab 133 \tab 24 \tab 312 \tab 13 \tab 20 \tab 115 \tab 559 \tab 53 \tab 36 \tab 69 \tab 29 \tab 47 \tab 124 \tab 10 \tab 30 \tab 46 \tab 44 \tab 27 \tab 95 \tab 31 \tab 44 \tab 77 \tab 25 \tab 72 \tab 13 \tab 22 \tab 273 \tab 113 \tab 294 \tab 65 \tab 143 \tab 23 \tab 288 \tab 40 \tab 94 \tab 282 \tab 875 \tab 150 \tab 84 \tab 210 \tab 345 \tab 779 \tab 170 \tab 119 \tab 673 \tab 993 \tab 1430 \tab 790 \tab 2702 \tab 48335 \tab 20 \tab 34 \tab 28 \tab 22 \tab 39 \tab 32 \tab 12 \tab 8 \tab 24 \tab 4 \tab 36 \tab 8 \tab 18 \tab 6 \tab 38 \tab 17 \tab 9 \tab 64 \tab 32 \tab 14 \tab 5 \tab 12 \tab 10 \tab 15 \tab 28 \tab 18 \tab 23 \tab 6 \tab 13 \tab 9 \tab 21 \tab 4 \tab 5 \tab 6 \tab 6 \tab 19 \tab 6 \tab 23 \tab 16 \tab 16 \tab 5 \tab 17 \tab 76 \tab 13 \tab 6 \tab 8 \tab 28 \tab 27 \tab 6 \tab 18 \tab 2 \tab 13 \tab 20 \tab 4 \tab 6 \tab 4 \tab 15 \tab 12 \tab 12 \tab 9 \tab 12 \tab 7 \tab 15 \tab 13 \tab 18 \tab 3 \tab 8 \tab 22 \tab 11 \tab 5 \tab 6 \tab 6 \tab 4 \tab 2 \tab 1 \tab 27 \tab 13 \tab 4 \tab 6 \tab 6 \tab 3 \tab 4 \tab 6 \tab 4 \tab 4 \tab 5 \tab 12 \tab 2 \tab 4 \tab 3 \tab 4 \tab 2 \tab 6 \tab 3 \tab 5 \tab 2 \tab 7 \tab 8 \tab 2 \tab 9 \tab 7 \tab 3 \tab 2 \tab 18 \tab 6 \tab 1 \tab 7 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
