% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{dwelling}
\alias{dwelling}
\title{dwelling}
\source{
General Social Survey https://gss.norc.org
}
\description{
Type of structure
}
\details{
Question 1609. DWELLING TYPE.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab 2-family house, 2 units one above the other \tab 2-family house, 2 units side-by-side \tab apartment house (5 or more units, 3 stories or less) \tab apartment house (5 or more units, 4 stories or more) \tab apartment in a partly commercial structure \tab detached 3-4 family house \tab detached single family house \tab no answer \tab other \tab row house (3 or more units in an attached row) \tab trailer \tab don't know \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 73 \tab 88 \tab 197 \tab 59 \tab 8 \tab 36 \tab 1155 \tab 17 \tab 11 \tab 96 \tab 120 \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 53 \tab 44 \tab 150 \tab 49 \tab 12 \tab 32 \tab 1065 \tab 15 \tab 12 \tab 85 \tab 82 \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 49 \tab 66 \tab 165 \tab 48 \tab 12 \tab 34 \tab 935 \tab 12 \tab 17 \tab 49 \tab 86 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 45 \tab 49 \tab 188 \tab 48 \tab 17 \tab 28 \tab 935 \tab 19 \tab 15 \tab 72 \tab 118 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 54 \tab 39 \tab 153 \tab 45 \tab 13 \tab 26 \tab 936 \tab 20 \tab 19 \tab 70 \tab 95 \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 50 \tab 73 \tab 207 \tab 84 \tab 14 \tab 31 \tab 1140 \tab 14 \tab 27 \tab 83 \tab 96 \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 52 \tab 52 \tab 177 \tab 56 \tab 8 \tab 39 \tab 848 \tab 27 \tab 10 \tab 99 \tab 113 \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 41 \tab 44 \tab 136 \tab 41 \tab 5 \tab 24 \tab 1006 \tab 27 \tab 39 \tab 57 \tab 117 \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 39 \tab 45 \tab 154 \tab 59 \tab 9 \tab 27 \tab 879 \tab 12 \tab 23 \tab 40 \tab 84 \tab 1 \tab - \tab 1372 \cr
   1991 \tab - \tab 58 \tab 62 \tab 171 \tab 51 \tab 13 \tab 45 \tab 953 \tab 14 \tab 8 \tab 46 \tab 95 \tab 1 \tab - \tab 1517 \cr
   1993 \tab - \tab 37 \tab 45 \tab 189 \tab 57 \tab 4 \tab 31 \tab 1043 \tab 1 \tab 13 \tab 78 \tab 108 \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 86 \tab 113 \tab 362 \tab 150 \tab 16 \tab 56 \tab 1790 \tab 15 \tab 19 \tab 159 \tab 195 \tab 31 \tab - \tab 2992 \cr
   1996 \tab - \tab 77 \tab 100 \tab 361 \tab 155 \tab 14 \tab 54 \tab 1700 \tab 30 \tab 30 \tab 153 \tab 218 \tab 12 \tab - \tab 2904 \cr
   1998 \tab - \tab 66 \tab 111 \tab 451 \tab 158 \tab 18 \tab 55 \tab 1609 \tab 22 \tab 26 \tab 139 \tab 173 \tab 4 \tab - \tab 2832 \cr
   2000 \tab - \tab 72 \tab 98 \tab 384 \tab 169 \tab 8 \tab 71 \tab 1574 \tab 57 \tab 20 \tab 167 \tab 193 \tab 4 \tab - \tab 2817 \cr
   2002 \tab - \tab 46 \tab 112 \tab 484 \tab 160 \tab 14 \tab 50 \tab 1492 \tab - \tab 52 \tab 129 \tab 197 \tab 29 \tab - \tab 2765 \cr
   2004 \tab - \tab 52 \tab 106 \tab 334 \tab 59 \tab 12 \tab 54 \tab 1841 \tab 9 \tab 23 \tab 117 \tab 158 \tab 47 \tab - \tab 2812 \cr
   2006 \tab - \tab 76 \tab 152 \tab 549 \tab 152 \tab 25 \tab 76 \tab 2853 \tab 7 \tab 65 \tab 220 \tab 299 \tab 36 \tab - \tab 4510 \cr
   2008 \tab - \tab 25 \tab 62 \tab 224 \tab 73 \tab 8 \tab 34 \tab 1336 \tab 3 \tab 23 \tab 89 \tab 118 \tab 28 \tab - \tab 2023 \cr
   2010 \tab - \tab 27 \tab 97 \tab 270 \tab 69 \tab 10 \tab 30 \tab 1244 \tab 8 \tab 25 \tab 102 \tab 107 \tab 55 \tab - \tab 2044 \cr
   2012 \tab - \tab 53 \tab 61 \tab 246 \tab 78 \tab 7 \tab 48 \tab 1269 \tab 6 \tab 4 \tab 74 \tab 128 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 30 \tab 68 \tab 312 \tab 125 \tab 11 \tab 29 \tab 1525 \tab 6 \tab 23 \tab 102 \tab 116 \tab 191 \tab - \tab 2538 \cr
   2016 \tab - \tab 62 \tab 104 \tab 400 \tab 119 \tab 9 \tab 38 \tab 1840 \tab 2 \tab 14 \tab 129 \tab 150 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 26 \tab 87 \tab 307 \tab 98 \tab 7 \tab 32 \tab 1519 \tab 1 \tab 6 \tab 127 \tab 138 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 1642 \tab 23 \tab 58 \tab 278 \tab 99 \tab 10 \tab 18 \tab 1069 \tab - \tab 26 \tab 90 \tab 113 \tab 118 \tab - \tab 3544 \cr
   Total \tab 13762 \tab 1272 \tab 1936 \tab 6849 \tab 2261 \tab 284 \tab 998 \tab 33556 \tab 344 \tab 550 \tab 2572 \tab 3417 \tab 557 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} trailer
\item \code{2} detached single family house
\item \code{3} 2-family house, 2 units side-by-side
\item \code{4} 2-family house, 2 units one above the other
\item \code{5} detached 3-4 family house
\item \code{6} row house (3 or more units in an attached row)
\item \code{7} apartment house (5 or more units, 3 stories or less)
\item \code{8} apartment house (5 or more units, 4 stories or more)
\item \code{9} apartment in a partly commercial structure
\item \code{10} other
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
