% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_15.R
\name{earnrs}
\alias{earnrs}
\title{How many in family earned money}
\source{
General Social Survey https://gss.norc.org
}
\description{
earnrs
}
\details{
Question 36. Just thinking about your family now -- those people in the househol who are related to you -- how many persons in the family, including yourself, earned any money last year from any job or employment?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllll}{
   year \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab no answer \tab 7 \tab eight or more \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 186 \tab 682 \tab 506 \tab 136 \tab 52 \tab 14 \tab 4 \tab 33 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 162 \tab 608 \tab 506 \tab 138 \tab 63 \tab 11 \tab 2 \tab 10 \tab 3 \tab 1 \tab - \tab - \tab 1504 \cr
   1974 \tab 192 \tab 595 \tab 477 \tab 132 \tab 50 \tab 15 \tab 7 \tab 15 \tab 1 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 201 \tab 593 \tab 483 \tab 121 \tab 54 \tab 25 \tab 5 \tab 8 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 233 \tab 627 \tab 449 \tab 118 \tab 45 \tab 17 \tab 1 \tab 8 \tab - \tab 1 \tab - \tab - \tab 1499 \cr
   1977 \tab 214 \tab 613 \tab 494 \tab 125 \tab 60 \tab 18 \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 229 \tab 629 \tab 479 \tab 104 \tab 54 \tab 22 \tab 3 \tab 8 \tab 2 \tab 2 \tab - \tab - \tab 1532 \cr
   1980 \tab 249 \tab 544 \tab 483 \tab 98 \tab 53 \tab 11 \tab 7 \tab 19 \tab 3 \tab 1 \tab - \tab - \tab 1468 \cr
   1982 \tab 291 \tab 752 \tab 577 \tab 135 \tab 38 \tab 10 \tab 4 \tab 53 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 234 \tab 621 \tab 544 \tab 123 \tab 41 \tab 16 \tab 2 \tab 17 \tab 1 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 233 \tab 611 \tab 441 \tab 111 \tab 42 \tab 13 \tab 4 \tab 18 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 276 \tab 578 \tab 499 \tab 99 \tab 55 \tab 11 \tab 3 \tab 11 \tab 2 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 297 \tab 506 \tab 478 \tab 107 \tab 41 \tab 17 \tab 2 \tab 20 \tab 2 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 291 \tab 709 \tab 585 \tab 148 \tab 53 \tab 15 \tab 2 \tab 16 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 275 \tab 560 \tab 452 \tab 137 \tab 34 \tab 9 \tab 4 \tab 10 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 286 \tab 553 \tab 475 \tab 133 \tab 52 \tab 20 \tab 4 \tab 12 \tab 1 \tab 1 \tab - \tab - \tab 1537 \cr
   1990 \tab 248 \tab 508 \tab 456 \tab 101 \tab 40 \tab 7 \tab 3 \tab 8 \tab - \tab 1 \tab - \tab - \tab 1372 \cr
   1991 \tab 320 \tab 581 \tab 459 \tab 100 \tab 42 \tab 4 \tab - \tab 11 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 298 \tab 577 \tab 556 \tab 114 \tab 32 \tab 6 \tab 2 \tab 20 \tab 1 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 493 \tab 1190 \tab 981 \tab 189 \tab 78 \tab 6 \tab - \tab 55 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 432 \tab 1213 \tab 920 \tab 186 \tab 70 \tab 15 \tab 5 \tab 63 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 429 \tab 1149 \tab 949 \tab 187 \tab 74 \tab 13 \tab 1 \tab 23 \tab 2 \tab 5 \tab - \tab - \tab 2832 \cr
   2000 \tab 471 \tab 1162 \tab 942 \tab 177 \tab 47 \tab 7 \tab 2 \tab 7 \tab - \tab 2 \tab - \tab - \tab 2817 \cr
   2002 \tab 466 \tab 1097 \tab 918 \tab 165 \tab 69 \tab 19 \tab 8 \tab 19 \tab 1 \tab 3 \tab - \tab - \tab 2765 \cr
   2004 \tab 394 \tab 1131 \tab 996 \tab 179 \tab 78 \tab 13 \tab 4 \tab 7 \tab 3 \tab 7 \tab - \tab - \tab 2812 \cr
   2006 \tab 645 \tab 1890 \tab 1502 \tab 290 \tab 121 \tab 37 \tab 4 \tab 21 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 328 \tab 791 \tab 668 \tab 166 \tab 44 \tab 18 \tab 4 \tab 3 \tab - \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab 354 \tab 825 \tab 632 \tab 143 \tab 55 \tab 13 \tab 5 \tab 14 \tab 2 \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab 369 \tab 770 \tab 637 \tab 119 \tab 56 \tab 8 \tab 2 \tab 12 \tab - \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab 467 \tab 1023 \tab 797 \tab 171 \tab 49 \tab 13 \tab 3 \tab 14 \tab 1 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 571 \tab 1104 \tab 891 \tab 180 \tab 82 \tab 17 \tab 1 \tab 19 \tab 1 \tab 1 \tab - \tab - \tab 2867 \cr
   2018 \tab 423 \tab 891 \tab 762 \tab 181 \tab 65 \tab 10 \tab 1 \tab 12 \tab 2 \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab 832 \tab 1421 \tab 1216 \tab 263 \tab 101 \tab 39 \tab 9 \tab - \tab 7 \tab 7 \tab 58 \tab 79 \tab 4032 \cr
   2022 \tab 736 \tab 1392 \tab 971 \tab 215 \tab 105 \tab 34 \tab 10 \tab 29 \tab 3 \tab 6 \tab 6 \tab 37 \tab 3544 \cr
   Total \tab 12125 \tab 28496 \tab 23181 \tab 5091 \tab 1995 \tab 523 \tab 120 \tab 599 \tab 38 \tab 42 \tab 64 \tab 116 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{8} eight or more
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
