% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_10.R
\name{educ}
\alias{educ}
\title{educ}
\source{
General Social Survey https://gss.norc.org
}
\description{
Highest year of school completed
}
\details{
Question 15. What is the highest grade in elementary school or high school that you finished and got credit for?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllll}{
   year \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab no formal schooling \tab don't know \tab Total \cr
   1972 \tab 4 \tab 116 \tab 119 \tab 507 \tab 92 \tab 124 \tab 53 \tab 119 \tab 26 \tab 22 \tab 12 \tab 13 \tab 12 \tab 19 \tab 19 \tab 30 \tab 38 \tab 48 \tab 142 \tab 79 \tab 5 \tab 14 \tab - \tab 1613 \cr
   1973 \tab 4 \tab 97 \tab 108 \tab 485 \tab 92 \tab 126 \tab 43 \tab 130 \tab 30 \tab 13 \tab 17 \tab 7 \tab 13 \tab 14 \tab 18 \tab 21 \tab 30 \tab 42 \tab 135 \tab 67 \tab 5 \tab 7 \tab - \tab 1504 \cr
   1974 \tab 1 \tab 107 \tab 90 \tab 491 \tab 104 \tab 109 \tab 43 \tab 128 \tab 39 \tab 28 \tab 9 \tab 1 \tab 17 \tab 9 \tab 13 \tab 17 \tab 30 \tab 45 \tab 132 \tab 62 \tab 3 \tab 6 \tab - \tab 1484 \cr
   1975 \tab 2 \tab 101 \tab 126 \tab 499 \tab 115 \tab 101 \tab 34 \tab 132 \tab 24 \tab 23 \tab 9 \tab 3 \tab 12 \tab 8 \tab 7 \tab 25 \tab 28 \tab 42 \tab 124 \tab 65 \tab 2 \tab 7 \tab 1 \tab 1490 \cr
   1976 \tab - \tab 110 \tab 93 \tab 510 \tab 94 \tab 103 \tab 36 \tab 134 \tab 31 \tab 23 \tab 17 \tab 4 \tab 14 \tab 15 \tab 19 \tab 19 \tab 24 \tab 46 \tab 128 \tab 69 \tab 4 \tab 4 \tab 2 \tab 1499 \cr
   1977 \tab - \tab 106 \tab 114 \tab 507 \tab 90 \tab 103 \tab 44 \tab 128 \tab 32 \tab 33 \tab 11 \tab 1 \tab 15 \tab 11 \tab 20 \tab 24 \tab 37 \tab 48 \tab 120 \tab 72 \tab 3 \tab 4 \tab 7 \tab 1530 \cr
   1978 \tab 4 \tab 96 \tab 90 \tab 541 \tab 105 \tab 120 \tab 55 \tab 119 \tab 34 \tab 34 \tab 17 \tab 1 \tab 16 \tab 8 \tab 20 \tab 13 \tab 19 \tab 39 \tab 126 \tab 63 \tab 3 \tab 6 \tab 3 \tab 1532 \cr
   1980 \tab 1 \tab 91 \tab 83 \tab 503 \tab 97 \tab 109 \tab 48 \tab 134 \tab 37 \tab 38 \tab 14 \tab 6 \tab 18 \tab 7 \tab 13 \tab 18 \tab 27 \tab 38 \tab 101 \tab 79 \tab 1 \tab 1 \tab 4 \tab 1468 \cr
   1982 \tab 2 \tab 106 \tab 140 \tab 620 \tab 142 \tab 159 \tab 66 \tab 146 \tab 33 \tab 36 \tab 25 \tab 5 \tab 16 \tab 15 \tab 8 \tab 21 \tab 48 \tab 51 \tab 119 \tab 87 \tab 4 \tab 7 \tab 4 \tab 1860 \cr
   1983 \tab 1 \tab 73 \tab 127 \tab 547 \tab 97 \tab 155 \tab 58 \tab 179 \tab 39 \tab 40 \tab 16 \tab 4 \tab 18 \tab 10 \tab 12 \tab 12 \tab 18 \tab 34 \tab 90 \tab 62 \tab - \tab 5 \tab 2 \tab 1599 \cr
   1984 \tab 3 \tab 99 \tab 91 \tab 491 \tab 108 \tab 121 \tab 74 \tab 149 \tab 40 \tab 31 \tab 20 \tab 4 \tab 25 \tab 9 \tab 11 \tab 11 \tab 19 \tab 26 \tab 77 \tab 58 \tab 3 \tab 3 \tab - \tab 1473 \cr
   1985 \tab 2 \tab 86 \tab 100 \tab 511 \tab 123 \tab 133 \tab 74 \tab 141 \tab 46 \tab 44 \tab 12 \tab 6 \tab 32 \tab 9 \tab 8 \tab 9 \tab 19 \tab 27 \tab 86 \tab 63 \tab - \tab 3 \tab - \tab 1534 \cr
   1986 \tab - \tab 90 \tab 113 \tab 488 \tab 103 \tab 127 \tab 50 \tab 163 \tab 41 \tab 36 \tab 17 \tab 2 \tab 25 \tab 9 \tab 8 \tab 13 \tab 24 \tab 23 \tab 78 \tab 53 \tab 1 \tab 6 \tab - \tab 1470 \cr
   1987 \tab 2 \tab 94 \tab 98 \tab 588 \tab 149 \tab 182 \tab 66 \tab 182 \tab 62 \tab 48 \tab 16 \tab 4 \tab 27 \tab 16 \tab 10 \tab 19 \tab 30 \tab 32 \tab 97 \tab 78 \tab 6 \tab 9 \tab 4 \tab 1819 \cr
   1988 \tab 1 \tab 71 \tab 104 \tab 455 \tab 124 \tab 158 \tab 66 \tab 164 \tab 32 \tab 42 \tab 18 \tab 1 \tab 28 \tab 6 \tab 12 \tab 12 \tab 23 \tab 27 \tab 75 \tab 58 \tab - \tab 1 \tab 3 \tab 1481 \cr
   1989 \tab 2 \tab 70 \tab 95 \tab 489 \tab 125 \tab 184 \tab 76 \tab 152 \tab 40 \tab 61 \tab 18 \tab - \tab 27 \tab 4 \tab 9 \tab 9 \tab 12 \tab 21 \tab 82 \tab 51 \tab 5 \tab 3 \tab 2 \tab 1537 \cr
   1990 \tab 1 \tab 53 \tab 81 \tab 436 \tab 106 \tab 167 \tab 62 \tab 163 \tab 44 \tab 49 \tab 13 \tab 5 \tab 28 \tab 1 \tab 9 \tab 5 \tab 19 \tab 14 \tab 58 \tab 52 \tab - \tab 4 \tab 2 \tab 1372 \cr
   1991 \tab - \tab 73 \tab 85 \tab 461 \tab 130 \tab 175 \tab 73 \tab 194 \tab 43 \tab 45 \tab 22 \tab - \tab 30 \tab 5 \tab 5 \tab 6 \tab 12 \tab 25 \tab 68 \tab 56 \tab 7 \tab 2 \tab - \tab 1517 \cr
   1993 \tab - \tab 58 \tab 86 \tab 477 \tab 154 \tab 177 \tab 75 \tab 224 \tab 50 \tab 73 \tab 25 \tab 4 \tab 27 \tab - \tab 7 \tab 7 \tab 20 \tab 27 \tab 61 \tab 47 \tab - \tab 3 \tab 4 \tab 1606 \cr
   1994 \tab 1 \tab 102 \tab 168 \tab 929 \tab 277 \tab 321 \tab 146 \tab 433 \tab 97 \tab 119 \tab 46 \tab 3 \tab 64 \tab 6 \tab 12 \tab 15 \tab 19 \tab 29 \tab 109 \tab 85 \tab 4 \tab 4 \tab 3 \tab 2992 \cr
   1996 \tab - \tab 121 \tab 155 \tab 849 \tab 296 \tab 339 \tab 153 \tab 392 \tab 120 \tab 125 \tab 42 \tab - \tab 86 \tab 8 \tab 6 \tab 13 \tab 14 \tab 23 \tab 80 \tab 69 \tab 1 \tab 4 \tab 8 \tab 2904 \cr
   1998 \tab - \tab 113 \tab 138 \tab 851 \tab 270 \tab 350 \tab 146 \tab 412 \tab 86 \tab 109 \tab 41 \tab 5 \tab 69 \tab 10 \tab 9 \tab 8 \tab 23 \tab 21 \tab 82 \tab 75 \tab 4 \tab 2 \tab 8 \tab 2832 \cr
   2000 \tab - \tab 106 \tab 153 \tab 823 \tab 290 \tab 374 \tab 127 \tab 393 \tab 90 \tab 114 \tab 39 \tab 3 \tab 68 \tab 6 \tab 6 \tab 9 \tab 22 \tab 20 \tab 95 \tab 69 \tab 9 \tab 1 \tab - \tab 2817 \cr
   2002 \tab 2 \tab 88 \tab 137 \tab 818 \tab 265 \tab 374 \tab 157 \tab 377 \tab 93 \tab 128 \tab 43 \tab 15 \tab 70 \tab 2 \tab 6 \tab 8 \tab 25 \tab 14 \tab 66 \tab 60 \tab 7 \tab 5 \tab 5 \tab 2765 \cr
   2004 \tab 1 \tab 72 \tab 156 \tab 734 \tab 265 \tab 403 \tab 174 \tab 450 \tab 106 \tab 138 \tab 51 \tab 7 \tab 93 \tab 3 \tab 2 \tab 4 \tab 17 \tab 17 \tab 54 \tab 59 \tab 2 \tab 4 \tab - \tab 2812 \cr
   2006 \tab 4 \tab 152 \tab 215 \tab 1204 \tab 422 \tab 628 \tab 212 \tab 687 \tab 167 \tab 208 \tab 78 \tab 28 \tab 112 \tab 13 \tab 11 \tab 23 \tab 69 \tab 32 \tab 85 \tab 127 \tab 9 \tab 22 \tab 2 \tab 4510 \cr
   2008 \tab - \tab 59 \tab 122 \tab 581 \tab 158 \tab 258 \tab 94 \tab 322 \tab 51 \tab 124 \tab 38 \tab 2 \tab 52 \tab 5 \tab 10 \tab 5 \tab 29 \tab 14 \tab 35 \tab 53 \tab 2 \tab 6 \tab 3 \tab 2023 \cr
   2010 \tab 1 \tab 66 \tab 108 \tab 558 \tab 186 \tab 229 \tab 109 \tab 334 \tab 71 \tab 101 \tab 33 \tab 5 \tab 72 \tab 4 \tab 9 \tab 6 \tab 30 \tab 12 \tab 45 \tab 55 \tab 4 \tab 5 \tab 1 \tab 2044 \cr
   2012 \tab 2 \tab 59 \tab 101 \tab 540 \tab 163 \tab 261 \tab 99 \tab 307 \tab 80 \tab 92 \tab 41 \tab 3 \tab 71 \tab 6 \tab 10 \tab 4 \tab 27 \tab 8 \tab 48 \tab 47 \tab 1 \tab 3 \tab 1 \tab 1974 \cr
   2014 \tab 1 \tab 90 \tab 134 \tab 689 \tab 193 \tab 318 \tab 133 \tab 452 \tab 78 \tab 139 \tab 52 \tab 3 \tab 100 \tab 6 \tab 3 \tab 9 \tab 25 \tab 8 \tab 43 \tab 53 \tab - \tab 8 \tab 1 \tab 2538 \cr
   2016 \tab 3 \tab 90 \tab 118 \tab 824 \tab 242 \tab 359 \tab 137 \tab 485 \tab 108 \tab 149 \tab 63 \tab 3 \tab 110 \tab 3 \tab 2 \tab 4 \tab 31 \tab 18 \tab 48 \tab 59 \tab 8 \tab 2 \tab 1 \tab 2867 \cr
   2018 \tab 2 \tab 65 \tab 95 \tab 657 \tab 183 \tab 313 \tab 127 \tab 430 \tab 97 \tab 119 \tab 45 \tab 4 \tab 72 \tab 10 \tab 5 \tab 3 \tab 20 \tab 8 \tab 35 \tab 51 \tab 1 \tab 4 \tab 2 \tab 2348 \cr
   2021 \tab 1 \tab 52 \tab 83 \tab 829 \tab 277 \tab 542 \tab 208 \tab 942 \tab 258 \tab 351 \tab 113 \tab 2 \tab 216 \tab 3 \tab 1 \tab 2 \tab 15 \tab 5 \tab 25 \tab 32 \tab 48 \tab 9 \tab 18 \tab 4032 \cr
   2022 \tab 1 \tab 78 \tab 116 \tab 909 \tab 268 \tab 506 \tab 189 \tab 697 \tab 167 \tab 210 \tab 79 \tab 4 \tab 148 \tab 8 \tab 6 \tab 6 \tab 23 \tab 12 \tab 37 \tab 57 \tab 20 \tab 3 \tab - \tab 3544 \cr
   Total \tab 49 \tab 3010 \tab 3942 \tab 21401 \tab 5905 \tab 8208 \tab 3307 \tab 9994 \tab 2392 \tab 2945 \tab 1112 \tab 158 \tab 1803 \tab 268 \tab 326 \tab 410 \tab 866 \tab 896 \tab 2786 \tab 2172 \tab 172 \tab 177 \tab 91 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} no formal schooling
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
