% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_600.R
\name{egypt}
\alias{egypt}
\title{egypt}
\source{
General Social Survey https://gss.norc.org
}
\description{
Liking for egypt
}
\details{
Question 103. You will notice that the boxes on this card go from the highest position of "plus 5" for a country which you like very much, to the lowest position of "minus 5" for a country you dislike very much. How far up the scale or how far down the scale would you rate the following countries? h. Egypt
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab not available in this year \tab +1 \tab +2 \tab +3 \tab +4 \tab +5 \tab -1 \tab -2 \tab -3 \tab -4 \tab -5 \tab dk, na, iap \tab iap \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 284 \tab 154 \tab 140 \tab 81 \tab 56 \tab 198 \tab 125 \tab 101 \tab 79 \tab 139 \tab 127 \tab - \tab 1484 \cr
   1975 \tab - \tab 309 \tab 150 \tab 113 \tab 47 \tab 43 \tab 224 \tab 125 \tab 121 \tab 80 \tab 129 \tab 149 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 335 \tab 168 \tab 148 \tab 63 \tab 72 \tab 254 \tab 92 \tab 75 \tab 55 \tab 82 \tab 186 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 386 \tab 244 \tab 280 \tab 167 \tab 117 \tab 180 \tab 79 \tab 80 \tab 45 \tab 73 \tab 209 \tab - \tab 1860 \cr
   1983 \tab - \tab 387 \tab 206 \tab 201 \tab 106 \tab 85 \tab 205 \tab 97 \tab 89 \tab 43 \tab 55 \tab 125 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 386 \tab 201 \tab 200 \tab 97 \tab 70 \tab 174 \tab 95 \tab 81 \tab 36 \tab 55 \tab 139 \tab - \tab 1534 \cr
   1986 \tab - \tab 359 \tab 210 \tab 143 \tab 78 \tab 62 \tab 210 \tab 98 \tab 88 \tab 39 \tab 89 \tab 94 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 228 \tab 98 \tab 107 \tab 50 \tab 58 \tab 137 \tab 63 \tab 58 \tab 20 \tab 60 \tab 602 \tab - \tab 1481 \cr
   1989 \tab - \tab 228 \tab 118 \tab 88 \tab 52 \tab 36 \tab 147 \tab 76 \tab 45 \tab 24 \tab 76 \tab 647 \tab - \tab 1537 \cr
   1990 \tab - \tab 246 \tab 109 \tab 99 \tab 35 \tab 46 \tab 110 \tab 60 \tab 38 \tab 23 \tab 58 \tab 548 \tab - \tab 1372 \cr
   1991 \tab - \tab 240 \tab 130 \tab 168 \tab 65 \tab 78 \tab 102 \tab 44 \tab 32 \tab 20 \tab 51 \tab 587 \tab - \tab 1517 \cr
   1993 \tab - \tab 283 \tab 117 \tab 108 \tab 61 \tab 66 \tab 140 \tab 56 \tab 60 \tab 36 \tab 57 \tab 622 \tab - \tab 1606 \cr
   1994 \tab - \tab 152 \tab 48 \tab 46 \tab 29 \tab 30 \tab 54 \tab 36 \tab 25 \tab 10 \tab 24 \tab 2538 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 47374 \tab 3823 \tab 1953 \tab 1841 \tab 931 \tab 819 \tab 2135 \tab 1046 \tab 893 \tab 510 \tab 948 \tab 6573 \tab 3544 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} +5
\item \code{1} +4
\item \code{2} +3
\item \code{3} +2
\item \code{4} +1
\item \code{5} -1
\item \code{6} -2
\item \code{7} -3
\item \code{8} -4
\item \code{9} -5
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
