% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{eqwlth}
\alias{eqwlth}
\title{eqwlth}
\source{
General Social Survey https://gss.norc.org
}
\description{
Should govt reduce income differences
}
\details{
Question 74a. Some people think that the government in Washington ought to reduce the income differences between the rich and the poor, perhaps by raising the taxes of wealthy families or by giving income assistance to the poor. Others think that the government should not concern itself with reducing this income difference between the rich and the poor. Here is a card with a scale from 1 to 7. Think of a score of 1 as meaning that the government ought to reduce the income differences between rich and poor, and a score of 7 meaning that the government should not concern itself with reducing income differences. What score between 1 and 7 comes closest to the way you feel?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab don't know \tab no answer \tab the government should not concern itself with reducing income differences \tab the government should reduce income differences \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 771 \tab 82 \tab 130 \tab 157 \tab 81 \tab 57 \tab 11 \tab 3 \tab 94 \tab 146 \tab - \tab 1532 \cr
   1980 \tab - \tab 136 \tab 238 \tab 290 \tab 180 \tab 106 \tab 40 \tab 5 \tab 227 \tab 246 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 171 \tab 256 \tab 278 \tab 177 \tab 131 \tab 38 \tab 3 \tab 230 \tab 315 \tab - \tab 1599 \cr
   1984 \tab - \tab 179 \tab 223 \tab 251 \tab 191 \tab 117 \tab 24 \tab 11 \tab 171 \tab 306 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 130 \tab 247 \tab 302 \tab 160 \tab 89 \tab 19 \tab 3 \tab 182 \tab 338 \tab - \tab 1470 \cr
   1987 \tab - \tab 160 \tab 307 \tab 369 \tab 218 \tab 103 \tab 25 \tab 8 \tab 221 \tab 408 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 95 \tab 180 \tab 197 \tab 121 \tab 75 \tab 21 \tab 3 \tab 112 \tab 193 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 132 \tab 199 \tab 206 \tab 116 \tab 72 \tab 18 \tab 2 \tab 108 \tab 182 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 107 \tab 157 \tab 190 \tab 80 \tab 56 \tab 25 \tab 3 \tab 93 \tab 188 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 126 \tab 173 \tab 200 \tab 123 \tab 75 \tab 27 \tab 3 \tab 89 \tab 201 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 122 \tab 201 \tab 187 \tab 130 \tab 81 \tab 18 \tab 4 \tab 130 \tab 184 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 183 \tab 322 \tab 420 \tab 292 \tab 167 \tab 27 \tab 8 \tab 303 \tab 289 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 200 \tab 303 \tab 409 \tab 237 \tab 158 \tab 45 \tab 8 \tab 234 \tab 331 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 181 \tab 331 \tab 402 \tab 213 \tab 157 \tab 35 \tab 7 \tab 303 \tab 282 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 217 \tab 294 \tab 374 \tab 258 \tab 160 \tab 38 \tab 9 \tab 245 \tab 301 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 84 \tab 164 \tab 177 \tab 125 \tab 66 \tab 11 \tab 1 \tab 109 \tab 175 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 80 \tab 152 \tab 168 \tab 107 \tab 64 \tab 10 \tab 1 \tab 123 \tab 171 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 187 \tab 320 \tab 410 \tab 254 \tab 149 \tab 28 \tab 5 \tab 221 \tab 415 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 116 \tab 234 \tab 240 \tab 178 \tab 93 \tab 15 \tab 1 \tab 169 \tab 319 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 106 \tab 219 \tab 249 \tab 215 \tab 99 \tab 12 \tab 6 \tab 227 \tab 244 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 132 \tab 194 \tab 240 \tab 148 \tab 84 \tab 12 \tab 2 \tab 202 \tab 324 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 165 \tab 251 \tab 284 \tab 232 \tab 121 \tab 15 \tab 1 \tab 253 \tab 365 \tab - \tab 2538 \cr
   2016 \tab 911 \tab 248 \tab 297 \tab 328 \tab 255 \tab 147 \tab 15 \tab 5 \tab 210 \tab 451 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 154 \tab 272 \tab 260 \tab 221 \tab 119 \tab 21 \tab 3 \tab 166 \tab 347 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 242 \tab 325 \tab 389 \tab 249 \tab 153 \tab 2 \tab - \tab 421 \tab 882 \tab 9 \tab 4032 \cr
   2022 \tab 1173 \tab 198 \tab 322 \tab 337 \tab 237 \tab 127 \tab 12 \tab 4 \tab 347 \tab 783 \tab 4 \tab 3544 \cr
   Total \tab 32946 \tab 3933 \tab 6311 \tab 7314 \tab 4798 \tab 2826 \tab 564 \tab 109 \tab 5190 \tab 8386 \tab 13 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} the government should reduce income differences
\item \code{7} the government should not concern itself with reducing income differences
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
