% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_573.R
\name{eth1}
\alias{eth1}
\title{1st mentioned country of origin}
\source{
General Social Survey https://gss.norc.org
}
\description{
eth1
}
\details{
Question A. IF MORE THAN ONE COUNTRY NAMED:Which one of these countries do you feel closer to?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab uncodeable \tab africa (general) \tab american indian \tab american only \tab austria \tab belgium \tab china \tab czechoslovakia \tab denmark \tab don't know \tab england & wales \tab finland \tab france \tab french canada \tab germany \tab greece \tab hungary \tab india \tab ireland \tab italy \tab japan \tab lithuania \tab mexico \tab netherlands \tab no answer \tab non-span west indies \tab norway \tab other \tab other canada \tab other spanish \tab philippines \tab poland \tab portugal \tab puerto rico \tab romania \tab russia \tab scotland \tab spain \tab sweden \tab switzerland \tab west indies \tab yugoslavia \tab arabic \tab other asian \tab other european \tab iap \tab argentina \tab belize \tab brazil \tab colombia \tab congo \tab cuba \tab dominican republic \tab egypt \tab el salvador \tab ethiopia \tab haiti \tab jamaica \tab kenya \tab nigeria \tab other africa \tab other middle east \tab other north america \tab other oceania \tab pakistan \tab panama \tab samoa \tab skipped on web \tab south africa \tab south korea \tab sudan \tab thailand \tab uganda \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 90 \tab 47 \tab 14 \tab 8 \tab 3 \tab 6 \tab 13 \tab 13 \tab 170 \tab 199 \tab 3 \tab 36 \tab 24 \tab 250 \tab 7 \tab 14 \tab 3 \tab 165 \tab 68 \tab 6 \tab 3 \tab 28 \tab 41 \tab 5 \tab 6 \tab 28 \tab 11 \tab 7 \tab 7 \tab 4 \tab 50 \tab 3 \tab 13 \tab 2 \tab 22 \tab 53 \tab 16 \tab 17 \tab 10 \tab 1 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 67 \tab 41 \tab 15 \tab 10 \tab 4 \tab 5 \tab 23 \tab 10 \tab 162 \tab 238 \tab 5 \tab 55 \tab 27 \tab 277 \tab 4 \tab 6 \tab 1 \tab 148 \tab 86 \tab 6 \tab 2 \tab 36 \tab 36 \tab 4 \tab 1 \tab 34 \tab 18 \tab 6 \tab 9 \tab 2 \tab 40 \tab 2 \tab 9 \tab 5 \tab 18 \tab 51 \tab 20 \tab 35 \tab 5 \tab 1 \tab 6 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 95 \tab 31 \tab 17 \tab 14 \tab - \tab 4 \tab 18 \tab 12 \tab 160 \tab 200 \tab 3 \tab 40 \tab 16 \tab 252 \tab 6 \tab 6 \tab 5 \tab 171 \tab 82 \tab 5 \tab 3 \tab 31 \tab 29 \tab 1 \tab 7 \tab 36 \tab 14 \tab 11 \tab 4 \tab 2 \tab 59 \tab 3 \tab 12 \tab 3 \tab 29 \tab 32 \tab 19 \tab 24 \tab 8 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 284 \tab 85 \tab 36 \tab 10 \tab 2 \tab 8 \tab 10 \tab 9 \tab 254 \tab 190 \tab 5 \tab 39 \tab 22 \tab 289 \tab 4 \tab 5 \tab 6 \tab 147 \tab 65 \tab 3 \tab 5 \tab 34 \tab 22 \tab 10 \tab 8 \tab 22 \tab 18 \tab 9 \tab 7 \tab 2 \tab 42 \tab 9 \tab 5 \tab 2 \tab 21 \tab 62 \tab 17 \tab 28 \tab 7 \tab 3 \tab 9 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 89 \tab 46 \tab 26 \tab 9 \tab 1 \tab 6 \tab 17 \tab 8 \tab 169 \tab 188 \tab 1 \tab 36 \tab 16 \tab 284 \tab 8 \tab 9 \tab 7 \tab 167 \tab 53 \tab 4 \tab 5 \tab 42 \tab 23 \tab 9 \tab 2 \tab 25 \tab 18 \tab 13 \tab 7 \tab 1 \tab 36 \tab 1 \tab 15 \tab 2 \tab 23 \tab 52 \tab 19 \tab 28 \tab 6 \tab - \tab 7 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 87 \tab 48 \tab 13 \tab 9 \tab 2 \tab 9 \tab 18 \tab 12 \tab 125 \tab 230 \tab 2 \tab 39 \tab 17 \tab 330 \tab 5 \tab 8 \tab 4 \tab 153 \tab 77 \tab 6 \tab 2 \tab 37 \tab 29 \tab 13 \tab 5 \tab 38 \tab 16 \tab 9 \tab 11 \tab 5 \tab 41 \tab 4 \tab 9 \tab 2 \tab 29 \tab 60 \tab 11 \tab 11 \tab 7 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 107 \tab 51 \tab 6 \tab 7 \tab 4 \tab 4 \tab 21 \tab 13 \tab 146 \tab 185 \tab 5 \tab 24 \tab 23 \tab 240 \tab 10 \tab 4 \tab 3 \tab 139 \tab 53 \tab 2 \tab 9 \tab 32 \tab 20 \tab 8 \tab 8 \tab 22 \tab 13 \tab 9 \tab 9 \tab 1 \tab 41 \tab 1 \tab 12 \tab 4 \tab 32 \tab 52 \tab 10 \tab 25 \tab 8 \tab - \tab 6 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 117 \tab 43 \tab 16 \tab 13 \tab 3 \tab 3 \tab 18 \tab 13 \tab 152 \tab 214 \tab 2 \tab 49 \tab 26 \tab 240 \tab 5 \tab 9 \tab 2 \tab 171 \tab 74 \tab 5 \tab 3 \tab 43 \tab 20 \tab 6 \tab 7 \tab 27 \tab 20 \tab 10 \tab 6 \tab 4 \tab 43 \tab 3 \tab 13 \tab 2 \tab 27 \tab 47 \tab 15 \tab 23 \tab 13 \tab - \tab 9 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 105 \tab 58 \tab 22 \tab 8 \tab 1 \tab 7 \tab 14 \tab 11 \tab 175 \tab 189 \tab 5 \tab 40 \tab 15 \tab 304 \tab 8 \tab 17 \tab 7 \tab 166 \tab 65 \tab 3 \tab 5 \tab 28 \tab 39 \tab 4 \tab 11 \tab 19 \tab 18 \tab 8 \tab 10 \tab 8 \tab 47 \tab 5 \tab 11 \tab 1 \tab 37 \tab 73 \tab 24 \tab 25 \tab 7 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 231 \tab 81 \tab 74 \tab 23 \tab 3 \tab 9 \tab 47 \tab 21 \tab 272 \tab 368 \tab 5 \tab 71 \tab 42 \tab 602 \tab 13 \tab 13 \tab 14 \tab 321 \tab 122 \tab 6 \tab 12 \tab 68 \tab 50 \tab 7 \tab 16 \tab 49 \tab 12 \tab 16 \tab 20 \tab 12 \tab 72 \tab 8 \tab 20 \tab 7 \tab 41 \tab 123 \tab 32 \tab 37 \tab 12 \tab 4 \tab 5 \tab 8 \tab 17 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 195 \tab 94 \tab 65 \tab 21 \tab 3 \tab 12 \tab 38 \tab 13 \tab 354 \tab 348 \tab 8 \tab 76 \tab 27 \tab 527 \tab 15 \tab 18 \tab 13 \tab 264 \tab 133 \tab 4 \tab 6 \tab 72 \tab 42 \tab 53 \tab 12 \tab 38 \tab 26 \tab 10 \tab 24 \tab 11 \tab 62 \tab 10 \tab 25 \tab 5 \tab 43 \tab 105 \tab 35 \tab 43 \tab 11 \tab - \tab 13 \tab 4 \tab 17 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 216 \tab 82 \tab 67 \tab 12 \tab 6 \tab 24 \tab 33 \tab 17 \tab 259 \tab 321 \tab 5 \tab 83 \tab 45 \tab 529 \tab 11 \tab 24 \tab 9 \tab 291 \tab 135 \tab 6 \tab 9 \tab 75 \tab 42 \tab 16 \tab 19 \tab 49 \tab 27 \tab 13 \tab 31 \tab 20 \tab 82 \tab 9 \tab 33 \tab 3 \tab 24 \tab 94 \tab 26 \tab 34 \tab 8 \tab 1 \tab 14 \tab 3 \tab 14 \tab 11 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 234 \tab 127 \tab 55 \tab 18 \tab 4 \tab 8 \tab 27 \tab 18 \tab 230 \tab 320 \tab 8 \tab 60 \tab 40 \tab 456 \tab 19 \tab 8 \tab 17 \tab 297 \tab 132 \tab 8 \tab 9 \tab 92 \tab 36 \tab 37 \tab 14 \tab 55 \tab 31 \tab 13 \tab 30 \tab 27 \tab 79 \tab 8 \tab 37 \tab 3 \tab 53 \tab 94 \tab 21 \tab 40 \tab 10 \tab 6 \tab 9 \tab 2 \tab 17 \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 241 \tab 77 \tab 59 \tab 24 \tab 10 \tab 12 \tab 17 \tab 19 \tab 219 \tab 297 \tab 7 \tab 76 \tab 34 \tab 523 \tab 17 \tab 22 \tab 14 \tab 282 \tab 127 \tab 7 \tab 10 \tab 81 \tab 34 \tab 7 \tab 16 \tab 44 \tab 5 \tab 13 \tab 28 \tab 17 \tab 86 \tab 7 \tab 34 \tab 8 \tab 53 \tab 96 \tab 34 \tab 51 \tab 7 \tab 4 \tab 5 \tab 7 \tab 17 \tab 17 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 233 \tab 134 \tab 57 \tab 10 \tab 7 \tab 27 \tab 28 \tab 19 \tab 219 \tab 348 \tab 12 \tab 58 \tab 12 \tab 440 \tab 11 \tab 12 \tab 23 \tab 338 \tab 130 \tab 11 \tab 6 \tab 132 \tab 30 \tab 12 \tab 14 \tab 39 \tab 10 \tab 13 \tab 29 \tab 24 \tab 54 \tab 12 \tab 27 \tab 6 \tab 47 \tab 88 \tab 28 \tab 35 \tab 13 \tab 1 \tab 8 \tab 13 \tab 24 \tab 18 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 407 \tab 104 \tab 125 \tab 14 \tab 13 \tab 28 \tab 57 \tab 33 \tab 314 \tab 457 \tab 13 \tab 102 \tab 18 \tab 737 \tab 19 \tab 27 \tab 29 \tab 475 \tab 215 \tab 19 \tab 11 \tab 327 \tab 55 \tab 14 \tab 26 \tab 55 \tab 20 \tab 18 \tab 99 \tab 26 \tab 112 \tab 19 \tab 49 \tab 6 \tab 62 \tab 159 \tab 71 \tab 73 \tab 22 \tab 5 \tab 14 \tab 13 \tab 28 \tab 20 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 156 \tab 53 \tab 42 \tab 10 \tab 7 \tab 14 \tab 18 \tab 11 \tab 160 \tab 226 \tab 5 \tab 45 \tab 8 \tab 338 \tab 5 \tab 11 \tab 24 \tab 221 \tab 97 \tab 5 \tab 10 \tab 126 \tab 22 \tab 8 \tab 21 \tab 29 \tab 9 \tab 11 \tab 42 \tab 14 \tab 43 \tab 13 \tab 19 \tab 5 \tab 26 \tab 54 \tab 36 \tab 33 \tab 11 \tab 3 \tab 5 \tab 3 \tab 15 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 164 \tab 53 \tab 56 \tab 2 \tab 7 \tab 12 \tab 17 \tab 17 \tab 185 \tab 212 \tab 6 \tab 43 \tab 15 \tab 359 \tab 9 \tab 6 \tab 13 \tab 223 \tab 87 \tab 12 \tab 5 \tab 124 \tab 26 \tab 6 \tab 15 \tab 24 \tab 8 \tab 4 \tab 39 \tab 10 \tab 57 \tab 19 \tab 21 \tab 5 \tab 30 \tab 60 \tab 28 \tab 16 \tab 5 \tab 2 \tab 5 \tab 5 \tab 21 \tab 11 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 178 \tab 64 \tab 63 \tab 10 \tab 2 \tab 16 \tab 10 \tab 10 \tab 158 \tab 187 \tab 10 \tab 41 \tab 15 \tab 327 \tab 14 \tab 8 \tab 24 \tab 204 \tab 93 \tab 2 \tab 5 \tab 131 \tab 16 \tab 5 \tab 13 \tab 22 \tab 19 \tab 8 \tab 38 \tab 12 \tab 39 \tab 12 \tab 17 \tab 4 \tab 23 \tab 48 \tab 51 \tab 24 \tab 9 \tab 2 \tab 8 \tab 4 \tab 19 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 201 \tab 67 \tab 47 \tab 14 \tab 3 \tab 23 \tab 18 \tab 9 \tab 209 \tab 256 \tab 10 \tab 46 \tab 23 \tab 418 \tab 14 \tab 14 \tab 31 \tab 267 \tab 132 \tab 9 \tab 9 \tab 191 \tab 25 \tab 8 \tab 23 \tab 24 \tab 5 \tab 8 \tab 76 \tab 17 \tab 59 \tab 12 \tab 30 \tab 7 \tab 29 \tab 66 \tab 47 \tab 38 \tab 4 \tab 3 \tab 4 \tab 8 \tab 15 \tab 19 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 268 \tab 90 \tab 57 \tab 6 \tab 8 \tab 23 \tab 32 \tab 14 \tab 271 \tab 244 \tab 9 \tab 59 \tab 21 \tab 455 \tab 17 \tab 14 \tab 26 \tab 274 \tab 149 \tab 8 \tab 13 \tab 184 \tab 24 \tab 6 \tab 15 \tab 35 \tab 12 \tab 15 \tab 65 \tab 23 \tab 84 \tab 14 \tab 35 \tab 5 \tab 42 \tab 86 \tab 47 \tab 41 \tab 13 \tab 3 \tab 12 \tab 12 \tab 21 \tab 15 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 196 \tab 62 \tab 61 \tab 14 \tab 3 \tab 19 \tab 22 \tab 8 \tab 292 \tab 192 \tab 12 \tab 53 \tab 18 \tab 360 \tab 10 \tab 9 \tab 20 \tab 210 \tab 97 \tab 6 \tab 8 \tab 180 \tab 13 \tab 8 \tab 20 \tab 28 \tab 11 \tab 13 \tab 51 \tab 18 \tab 55 \tab 7 \tab 26 \tab 4 \tab 35 \tab 63 \tab 46 \tab 24 \tab 10 \tab 1 \tab 4 \tab 8 \tab 23 \tab 28 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab 29 \tab - \tab 2 \tab 5 \tab 1 \tab 10 \tab 7 \tab 17 \tab 144 \tab 3 \tab 37 \tab 4 \tab 136 \tab 4 \tab 2 \tab 1 \tab 136 \tab 31 \tab 2 \tab 9 \tab 6 \tab 11 \tab 2 \tab - \tab 18 \tab - \tab 6 \tab - \tab - \tab 25 \tab 8 \tab 4 \tab - \tab 11 \tab 66 \tab 25 \tab 23 \tab 5 \tab - \tab 2 \tab - \tab 1 \tab 11 \tab 2683 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 7 \tab 7 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3544 \cr
   Total \tab 15579 \tab 3961 \tab 1567 \tab 993 \tab 268 \tab 101 \tab 280 \tab 526 \tab 317 \tab 4672 \tab 5753 \tab 144 \tab 1208 \tab 508 \tab 8673 \tab 235 \tab 266 \tab 296 \tab 5230 \tab 2303 \tab 145 \tab 159 \tab 2100 \tab 685 \tab 249 \tab 279 \tab 760 \tab 341 \tab 243 \tab 642 \tab 260 \tab 1308 \tab 189 \tab 476 \tab 91 \tab 757 \tab 1684 \tab 678 \tab 728 \tab 211 \tab 45 \tab 158 \tab 110 \tab 249 \tab 191 \tab 6715 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 7 \tab 7 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} africa (general)
\item \code{2} austria
\item \code{3} french canada
\item \code{4} other canada
\item \code{5} china
\item \code{6} czechoslovakia
\item \code{7} denmark
\item \code{8} england & wales
\item \code{9} finland
\item \code{10} france
\item \code{11} germany
\item \code{12} greece
\item \code{13} hungary
\item \code{14} ireland
\item \code{15} italy
\item \code{16} japan
\item \code{17} mexico
\item \code{18} netherlands
\item \code{19} norway
\item \code{20} philippines
\item \code{21} poland
\item \code{22} puerto rico
\item \code{23} russia
\item \code{24} scotland
\item \code{25} spain
\item \code{26} sweden
\item \code{27} switzerland
\item \code{28} west indies
\item \code{29} other
\item \code{30} american indian
\item \code{31} india
\item \code{32} portugal
\item \code{33} lithuania
\item \code{34} yugoslavia
\item \code{35} romania
\item \code{36} belgium
\item \code{37} arabic
\item \code{38} other spanish
\item \code{39} non-span west indies
\item \code{40} other asian
\item \code{41} other european
\item \code{97} american only
\item \code{101} turkey
\item \code{202} algeria
\item \code{203} congo
\item \code{204} egypt
\item \code{205} ethiopia
\item \code{206} kenya
\item \code{207} nigeria
\item \code{208} south africa
\item \code{209} sudan
\item \code{210} tanzania
\item \code{211} uganda
\item \code{299} other africa
\item \code{301} south korea
\item \code{302} bangladesh
\item \code{303} indonesia
\item \code{304} pakistan
\item \code{306} thailand
\item \code{307} vietnam
\item \code{401} iran
\item \code{402} iraq
\item \code{403} israel
\item \code{404} jordan
\item \code{405} saudi arabia
\item \code{406} syria
\item \code{407} uae
\item \code{408} yemen
\item \code{499} other middle east
\item \code{501} argentina
\item \code{502} bolivia
\item \code{503} brazil
\item \code{504} chile
\item \code{505} colombia
\item \code{506} ecuador
\item \code{507} french guiana
\item \code{508} guyana
\item \code{509} paraguay
\item \code{510} peru
\item \code{511} suriname
\item \code{512} uruguay
\item \code{513} venezuela
\item \code{599} other south america
\item \code{601} belize
\item \code{602} costa rica
\item \code{603} el salvador
\item \code{604} guatemala
\item \code{605} honduras
\item \code{606} nicaragua
\item \code{607} panama
\item \code{699} other central america
\item \code{799} other north america
\item \code{801} cuba
\item \code{802} haiti
\item \code{803} dominican republic
\item \code{804} jamaica
\item \code{899} other caribbean
\item \code{901} australia
\item \code{902} papua new guinea
\item \code{903} new zealand
\item \code{904} samoa
\item \code{999} other oceania
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
