% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_574.R
\name{eth2}
\alias{eth2}
\title{eth2}
\source{
General Social Survey https://gss.norc.org
}
\description{
2nd mentioned country of origin
}
\details{
Question A. IF MORE THAN ONE COUNTRY NAMED:Which one of these countries do you feel closer to?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab uncodeable \tab africa (general) \tab american indian \tab american only \tab arabic \tab austria \tab belgium \tab czechoslovakia \tab denmark \tab england & wales \tab finland \tab france \tab french canada \tab germany \tab greece \tab hungary \tab india \tab ireland \tab italy \tab lithuania \tab mexico \tab netherlands \tab no answer \tab non-span west indies \tab norway \tab other \tab other canada \tab other spanish \tab poland \tab romania \tab russia \tab scotland \tab spain \tab sweden \tab switzerland \tab yugoslavia \tab japan \tab portugal \tab puerto rico \tab china \tab west indies \tab don't know \tab philippines \tab other asian \tab other european \tab iap \tab belize \tab congo \tab dominican republic \tab egypt \tab ethiopia \tab french guiana \tab kenya \tab nigeria \tab other africa \tab other south america \tab skipped on web \tab uganda \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 849 \tab 3 \tab 29 \tab 1 \tab 4 \tab 9 \tab 1 \tab 8 \tab 7 \tab 101 \tab 3 \tab 35 \tab 4 \tab 104 \tab 2 \tab 3 \tab 1 \tab 117 \tab 11 \tab 2 \tab 1 \tab 35 \tab 5 \tab 2 \tab 18 \tab 4 \tab 6 \tab 1 \tab 16 \tab 1 \tab 11 \tab 50 \tab 8 \tab 13 \tab 5 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 816 \tab 4 \tab 27 \tab - \tab - \tab 5 \tab 4 \tab 5 \tab 8 \tab 119 \tab 3 \tab 52 \tab 8 \tab 108 \tab 1 \tab 3 \tab - \tab 164 \tab 4 \tab 3 \tab 3 \tab 25 \tab 4 \tab 3 \tab 13 \tab 7 \tab 11 \tab 4 \tab 15 \tab 1 \tab 15 \tab 52 \tab 14 \tab 21 \tab 7 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 840 \tab 1 \tab 26 \tab 1 \tab 2 \tab 7 \tab 1 \tab 9 \tab 4 \tab 124 \tab - \tab 27 \tab 6 \tab 113 \tab 1 \tab 7 \tab 1 \tab 116 \tab 10 \tab 3 \tab 7 \tab 31 \tab 1 \tab 1 \tab 14 \tab 5 \tab 6 \tab 1 \tab 21 \tab - \tab 8 \tab 47 \tab 7 \tab 13 \tab 3 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1132 \tab 15 \tab 43 \tab 2 \tab 1 \tab 5 \tab 3 \tab 11 \tab 10 \tab 97 \tab 2 \tab 45 \tab 12 \tab 108 \tab - \tab 3 \tab 4 \tab 141 \tab 7 \tab 3 \tab 5 \tab 20 \tab 10 \tab 2 \tab 13 \tab 8 \tab 2 \tab 2 \tab 19 \tab 1 \tab 9 \tab 46 \tab 6 \tab 18 \tab 6 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 879 \tab 7 \tab 35 \tab 2 \tab 1 \tab 8 \tab - \tab 7 \tab 7 \tab 92 \tab - \tab 43 \tab 3 \tab 86 \tab - \tab 6 \tab 1 \tab 139 \tab 12 \tab 3 \tab 6 \tab 27 \tab 9 \tab - \tab 9 \tab 2 \tab 4 \tab 2 \tab 14 \tab 3 \tab 9 \tab 28 \tab 6 \tab 14 \tab 13 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 891 \tab 6 \tab 26 \tab 1 \tab 2 \tab 6 \tab - \tab 10 \tab 11 \tab 105 \tab 3 \tab 34 \tab 5 \tab 102 \tab 1 \tab 8 \tab 2 \tab 147 \tab 15 \tab 2 \tab 1 \tab 21 \tab 13 \tab - \tab 11 \tab 7 \tab 1 \tab 2 \tab 15 \tab 1 \tab 7 \tab 45 \tab 7 \tab 16 \tab 5 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 782 \tab 3 \tab 24 \tab - \tab - \tab 3 \tab 1 \tab 11 \tab 6 \tab 95 \tab 3 \tab 41 \tab 5 \tab 120 \tab 2 \tab 6 \tab 1 \tab 98 \tab 12 \tab 5 \tab 1 \tab 15 \tab 8 \tab 2 \tab 10 \tab 5 \tab 5 \tab 1 \tab 11 \tab 4 \tab 7 \tab 44 \tab 12 \tab 13 \tab 6 \tab 2 \tab - \tab 6 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 907 \tab 6 \tab 20 \tab - \tab 1 \tab 8 \tab - \tab 4 \tab 6 \tab 110 \tab 1 \tab 41 \tab 8 \tab 104 \tab 2 \tab 5 \tab 2 \tab 108 \tab 18 \tab 2 \tab 6 \tab 20 \tab 2 \tab - \tab 18 \tab 2 \tab 3 \tab 1 \tab 21 \tab - \tab 11 \tab 50 \tab 8 \tab 13 \tab 2 \tab 5 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 907 \tab 4 \tab 40 \tab 1 \tab 1 \tab 3 \tab 4 \tab 12 \tab 5 \tab 96 \tab 2 \tab 38 \tab 8 \tab 126 \tab 5 \tab 7 \tab - \tab 154 \tab 18 \tab 4 \tab 4 \tab 19 \tab - \tab 1 \tab 16 \tab 4 \tab 3 \tab 1 \tab 17 \tab 3 \tab 16 \tab 47 \tab 8 \tab 13 \tab 5 \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1443 \tab 12 \tab 61 \tab - \tab 5 \tab 17 \tab 1 \tab 17 \tab 13 \tab 216 \tab 4 \tab 66 \tab 13 \tab 202 \tab 6 \tab 8 \tab 5 \tab 258 \tab 32 \tab 3 \tab 9 \tab 36 \tab 7 \tab 6 \tab 24 \tab 7 \tab 13 \tab 2 \tab 31 \tab 2 \tab 34 \tab 95 \tab 17 \tab 29 \tab 15 \tab 3 \tab 3 \tab 3 \tab 3 \tab 1 \tab - \tab 261 \tab 3 \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1510 \tab 11 \tab 76 \tab 1 \tab 1 \tab 9 \tab 3 \tab 14 \tab 12 \tab 151 \tab 4 \tab 87 \tab 7 \tab 218 \tab 4 \tab 10 \tab 4 \tab 235 \tab 26 \tab 5 \tab 8 \tab 26 \tab 116 \tab 2 \tab 25 \tab 8 \tab 11 \tab 5 \tab 37 \tab 5 \tab 20 \tab 92 \tab 16 \tab 29 \tab 6 \tab 4 \tab 1 \tab 4 \tab 3 \tab 2 \tab - \tab 90 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1708 \tab 17 \tab 73 \tab 5 \tab 2 \tab 11 \tab 2 \tab 13 \tab 17 \tab 148 \tab 5 \tab 60 \tab 13 \tab 212 \tab 2 \tab 12 \tab 6 \tab 218 \tab 32 \tab 2 \tab 7 \tab 18 \tab 1 \tab 3 \tab 17 \tab 3 \tab 7 \tab 5 \tab 39 \tab 3 \tab 15 \tab 69 \tab 11 \tab 24 \tab 5 \tab 5 \tab 1 \tab 6 \tab 3 \tab 3 \tab - \tab 20 \tab 2 \tab 1 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1701 \tab 16 \tab 55 \tab 2 \tab 1 \tab 10 \tab 5 \tab 16 \tab 15 \tab 157 \tab 6 \tab 63 \tab 8 \tab 155 \tab 4 \tab 8 \tab 1 \tab 197 \tab 36 \tab 4 \tab 12 \tab 40 \tab 37 \tab 2 \tab 30 \tab 12 \tab 7 \tab 3 \tab 29 \tab 1 \tab 30 \tab 76 \tab 16 \tab 28 \tab 9 \tab 5 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 3 \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1287 \tab 22 \tab 95 \tab 12 \tab 1 \tab 16 \tab 2 \tab 28 \tab 17 \tab 197 \tab 3 \tab 70 \tab 18 \tab 226 \tab 5 \tab 10 \tab 4 \tab 268 \tab 40 \tab 2 \tab 17 \tab 36 \tab 10 \tab 1 \tab 23 \tab 4 \tab 8 \tab 17 \tab 47 \tab 8 \tab 34 \tab 110 \tab 36 \tab 38 \tab 9 \tab 3 \tab - \tab 9 \tab 4 \tab 3 \tab 1 \tab 8 \tab 2 \tab 4 \tab 10 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1571 \tab 11 \tab 64 \tab 3 \tab 3 \tab 13 \tab 4 \tab 16 \tab 22 \tab 178 \tab 1 \tab 57 \tab 13 \tab 224 \tab 7 \tab 13 \tab 2 \tab 224 \tab 33 \tab 4 \tab 16 \tab 25 \tab 11 \tab 7 \tab 13 \tab 4 \tab 7 \tab 7 \tab 35 \tab 6 \tab 19 \tab 86 \tab 30 \tab 26 \tab 7 \tab 5 \tab 1 \tab 4 \tab 6 \tab 3 \tab 2 \tab 13 \tab 3 \tab 7 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2621 \tab 15 \tab 118 \tab 5 \tab 5 \tab 9 \tab 8 \tab 26 \tab 32 \tab 279 \tab 10 \tab 98 \tab 14 \tab 309 \tab 3 \tab 7 \tab 2 \tab 344 \tab 53 \tab 10 \tab 21 \tab 56 \tab 14 \tab 5 \tab 34 \tab 7 \tab 9 \tab 15 \tab 55 \tab 5 \tab 35 \tab 131 \tab 40 \tab 46 \tab 21 \tab 7 \tab 1 \tab 7 \tab 8 \tab - \tab 2 \tab 8 \tab 2 \tab 4 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1248 \tab 6 \tab 45 \tab 1 \tab 3 \tab 7 \tab 4 \tab 12 \tab 16 \tab 114 \tab 1 \tab 36 \tab 6 \tab 138 \tab 3 \tab 9 \tab - \tab 156 \tab 19 \tab 3 \tab 9 \tab 18 \tab 8 \tab 1 \tab 12 \tab - \tab 2 \tab 1 \tab 32 \tab 1 \tab 10 \tab 58 \tab 13 \tab 19 \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1170 \tab 11 \tab 35 \tab 6 \tab 2 \tab 5 \tab 3 \tab 12 \tab 8 \tab 126 \tab 3 \tab 54 \tab 7 \tab 146 \tab 1 \tab 5 \tab 2 \tab 157 \tab 20 \tab 4 \tab 13 \tab 27 \tab 6 \tab 1 \tab 20 \tab 2 \tab 11 \tab 9 \tab 20 \tab 4 \tab 15 \tab 71 \tab 10 \tab 20 \tab 12 \tab 3 \tab 2 \tab 4 \tab 4 \tab 3 \tab 2 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1257 \tab 9 \tab 35 \tab - \tab 2 \tab 9 \tab 2 \tab 9 \tab 6 \tab 78 \tab 4 \tab 42 \tab 10 \tab 130 \tab 1 \tab 7 \tab 1 \tab 128 \tab 23 \tab 3 \tab 10 \tab 17 \tab 5 \tab 2 \tab 14 \tab 4 \tab 2 \tab 9 \tab 36 \tab 3 \tab 10 \tab 53 \tab 11 \tab 16 \tab 4 \tab 1 \tab 1 \tab 4 \tab 2 \tab 2 \tab 1 \tab - \tab 3 \tab 2 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1499 \tab 13 \tab 54 \tab 4 \tab 2 \tab 13 \tab 1 \tab 10 \tab 7 \tab 129 \tab 4 \tab 46 \tab 12 \tab 173 \tab 5 \tab 6 \tab 5 \tab 185 \tab 40 \tab - \tab 8 \tab 23 \tab 8 \tab 6 \tab 28 \tab 5 \tab 10 \tab 14 \tab 43 \tab 2 \tab 16 \tab 83 \tab 31 \tab 23 \tab 4 \tab 5 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 5 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1579 \tab 26 \tab 73 \tab 3 \tab 2 \tab 16 \tab 3 \tab 22 \tab 7 \tab 178 \tab 6 \tab 59 \tab 19 \tab 190 \tab 2 \tab 9 \tab 3 \tab 228 \tab 46 \tab 8 \tab 15 \tab 30 \tab 6 \tab 5 \tab 25 \tab 1 \tab 11 \tab 14 \tab 48 \tab 1 \tab 20 \tab 96 \tab 27 \tab 36 \tab 10 \tab 6 \tab 3 \tab 3 \tab 6 \tab 4 \tab 4 \tab - \tab - \tab 9 \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1371 \tab 13 \tab 38 \tab 5 \tab 1 \tab 8 \tab 4 \tab 11 \tab 11 \tab 133 \tab 6 \tab 49 \tab 15 \tab 141 \tab 1 \tab 5 \tab 3 \tab 190 \tab 32 \tab 3 \tab 14 \tab 20 \tab 8 \tab 6 \tab 11 \tab 3 \tab 10 \tab 9 \tab 40 \tab 3 \tab 22 \tab 60 \tab 23 \tab 21 \tab 9 \tab 7 \tab 4 \tab 3 \tab 7 \tab 2 \tab 3 \tab - \tab 2 \tab 3 \tab 18 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab 12 \tab - \tab - \tab 3 \tab 3 \tab 7 \tab 6 \tab 62 \tab 2 \tab 28 \tab 5 \tab 57 \tab 2 \tab 3 \tab - \tab 54 \tab 3 \tab 2 \tab 2 \tab 5 \tab 1 \tab - \tab 12 \tab - \tab 2 \tab - \tab 13 \tab 1 \tab 4 \tab 32 \tab 4 \tab 11 \tab 4 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 9 \tab 1 \tab - \tab 3 \tab 3163 \tab 1 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 4 \tab 2 \tab 3544 \cr
   Total \tab 43547 \tab 231 \tab 1104 \tab 55 \tab 42 \tab 200 \tab 59 \tab 290 \tab 253 \tab 3085 \tab 76 \tab 1171 \tab 219 \tab 3492 \tab 60 \tab 160 \tab 50 \tab 4026 \tab 542 \tab 80 \tab 195 \tab 590 \tab 290 \tab 58 \tab 410 \tab 104 \tab 151 \tab 125 \tab 654 \tab 59 \tab 377 \tab 1521 \tab 361 \tab 500 \tab 173 \tab 84 \tab 25 \tab 72 \tab 55 \tab 29 \tab 18 \tab 417 \tab 30 \tab 45 \tab 85 \tab 7195 \tab 1 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 4 \tab 2 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} africa (general)
\item \code{2} austria
\item \code{3} french canada
\item \code{4} other canada
\item \code{5} china
\item \code{6} czechoslovakia
\item \code{7} denmark
\item \code{8} england & wales
\item \code{9} finland
\item \code{10} france
\item \code{11} germany
\item \code{12} greece
\item \code{13} hungary
\item \code{14} ireland
\item \code{15} italy
\item \code{16} japan
\item \code{17} mexico
\item \code{18} netherlands
\item \code{19} norway
\item \code{20} philippines
\item \code{21} poland
\item \code{22} puerto rico
\item \code{23} russia
\item \code{24} scotland
\item \code{25} spain
\item \code{26} sweden
\item \code{27} switzerland
\item \code{28} west indies
\item \code{29} other
\item \code{30} american indian
\item \code{31} india
\item \code{32} portugal
\item \code{33} lithuania
\item \code{34} yugoslavia
\item \code{35} romania
\item \code{36} belgium
\item \code{37} arabic
\item \code{38} other spanish
\item \code{39} non-span west indies
\item \code{40} other asian
\item \code{41} other european
\item \code{97} american only
\item \code{101} turkey
\item \code{202} algeria
\item \code{203} congo
\item \code{204} egypt
\item \code{205} ethiopia
\item \code{206} kenya
\item \code{207} nigeria
\item \code{208} south africa
\item \code{209} sudan
\item \code{210} tanzania
\item \code{211} uganda
\item \code{299} other africa
\item \code{301} south korea
\item \code{302} bangladesh
\item \code{303} indonesia
\item \code{304} pakistan
\item \code{306} thailand
\item \code{307} vietnam
\item \code{401} iran
\item \code{402} iraq
\item \code{403} israel
\item \code{404} jordan
\item \code{405} saudi arabia
\item \code{406} syria
\item \code{407} uae
\item \code{408} yemen
\item \code{499} other middle east
\item \code{501} argentina
\item \code{502} bolivia
\item \code{503} brazil
\item \code{504} chile
\item \code{505} colombia
\item \code{506} ecuador
\item \code{507} french guiana
\item \code{508} guyana
\item \code{509} paraguay
\item \code{510} peru
\item \code{511} suriname
\item \code{512} uruguay
\item \code{513} venezuela
\item \code{599} other south america
\item \code{601} belize
\item \code{602} costa rica
\item \code{603} el salvador
\item \code{604} guatemala
\item \code{605} honduras
\item \code{606} nicaragua
\item \code{607} panama
\item \code{699} other central america
\item \code{799} other north america
\item \code{801} cuba
\item \code{802} haiti
\item \code{803} dominican republic
\item \code{804} jamaica
\item \code{899} other caribbean
\item \code{901} australia
\item \code{902} papua new guinea
\item \code{903} new zealand
\item \code{904} samoa
\item \code{999} other oceania
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
