% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_574.R
\name{eth3}
\alias{eth3}
\title{3rd mentioned country of origin}
\source{
General Social Survey https://gss.norc.org
}
\description{
eth3
}
\details{
Question A. IF MORE THAN ONE COUNTRY NAMED:Which one of these countries do you feel closer to?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab uncodeable \tab american indian \tab american only \tab czechoslovakia \tab denmark \tab england & wales \tab france \tab french canada \tab germany \tab hungary \tab ireland \tab italy \tab mexico \tab netherlands \tab no answer \tab norway \tab other \tab poland \tab puerto rico \tab romania \tab russia \tab scotland \tab spain \tab sweden \tab austria \tab belgium \tab finland \tab india \tab lithuania \tab non-span west indies \tab other canada \tab other spanish \tab philippines \tab yugoslavia \tab portugal \tab africa (general) \tab switzerland \tab west indies \tab arabic \tab japan \tab don't know \tab greece \tab other asian \tab other european \tab china \tab iap \tab cuba \tab dominican republic \tab iran \tab jamaica \tab kenya \tab skipped on web \tab south korea \tab tanzania \tab turkey \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1244 \tab 17 \tab 1 \tab 1 \tab 2 \tab 37 \tab 20 \tab 4 \tab 32 \tab 1 \tab 49 \tab 5 \tab 1 \tab 11 \tab 5 \tab 3 \tab 4 \tab 6 \tab 1 \tab 1 \tab 1 \tab 18 \tab 3 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1279 \tab 19 \tab - \tab 1 \tab 3 \tab 37 \tab 20 \tab 4 \tab 43 \tab 1 \tab 39 \tab 2 \tab 1 \tab 17 \tab 4 \tab 4 \tab 1 \tab 3 \tab - \tab - \tab - \tab 27 \tab - \tab 9 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1264 \tab 15 \tab - \tab 3 \tab 5 \tab 32 \tab 12 \tab 3 \tab 32 \tab - \tab 41 \tab 5 \tab 3 \tab 12 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 18 \tab 5 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1561 \tab 21 \tab 1 \tab 4 \tab 5 \tab 53 \tab 9 \tab 1 \tab 35 \tab 2 \tab 35 \tab 5 \tab 1 \tab 8 \tab 10 \tab 6 \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab 23 \tab 4 \tab 6 \tab 1 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 4 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1274 \tab 12 \tab 1 \tab 2 \tab 3 \tab 26 \tab 24 \tab - \tab 27 \tab 2 \tab 26 \tab 2 \tab 3 \tab 8 \tab 9 \tab 4 \tab - \tab 2 \tab - \tab - \tab 5 \tab 27 \tab 4 \tab 6 \tab 3 \tab - \tab - \tab 3 \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1312 \tab 20 \tab - \tab 3 \tab 6 \tab 29 \tab 15 \tab 3 \tab 36 \tab - \tab 33 \tab 9 \tab 1 \tab 12 \tab 13 \tab 3 \tab 1 \tab 5 \tab - \tab - \tab 2 \tab 22 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1149 \tab 16 \tab - \tab 3 \tab 1 \tab 39 \tab 20 \tab 4 \tab 32 \tab 2 \tab 36 \tab 6 \tab - \tab 11 \tab 8 \tab 2 \tab 3 \tab 8 \tab - \tab - \tab 4 \tab 15 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1305 \tab 19 \tab 1 \tab 5 \tab 6 \tab 26 \tab 14 \tab 4 \tab 30 \tab 1 \tab 31 \tab 3 \tab 1 \tab 14 \tab 2 \tab 4 \tab 1 \tab 8 \tab 1 \tab - \tab 4 \tab 16 \tab 3 \tab 5 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1376 \tab 24 \tab 1 \tab 1 \tab 2 \tab 30 \tab 20 \tab 1 \tab 36 \tab 1 \tab 40 \tab 2 \tab 1 \tab 12 \tab 3 \tab 6 \tab 1 \tab 4 \tab - \tab - \tab 5 \tab 22 \tab - \tab 7 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2249 \tab 41 \tab 1 \tab 6 \tab 5 \tab 61 \tab 35 \tab 3 \tab 77 \tab 4 \tab 84 \tab 8 \tab 5 \tab 7 \tab 9 \tab 7 \tab 4 \tab 14 \tab 1 \tab 2 \tab 7 \tab 62 \tab 9 \tab 10 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 4 \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 257 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1751 \tab 28 \tab 3 \tab 4 \tab 4 \tab 66 \tab 33 \tab 3 \tab 59 \tab 7 \tab 59 \tab 15 \tab 1 \tab 15 \tab 750 \tab 8 \tab 2 \tab 8 \tab 1 \tab - \tab 2 \tab 50 \tab 4 \tab 6 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 6 \tab - \tab 2 \tab 3 \tab 1 \tab 6 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2386 \tab 52 \tab 7 \tab 5 \tab 7 \tab 64 \tab 42 \tab 9 \tab 45 \tab 4 \tab 66 \tab 7 \tab 1 \tab 14 \tab - \tab 14 \tab 6 \tab 14 \tab - \tab - \tab 5 \tab 36 \tab 4 \tab 7 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 6 \tab 4 \tab - \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2357 \tab 28 \tab - \tab 5 \tab 7 \tab 57 \tab 29 \tab 4 \tab 67 \tab 2 \tab 74 \tab 16 \tab 1 \tab 15 \tab 37 \tab 8 \tab 3 \tab 19 \tab 1 \tab 1 \tab 11 \tab 42 \tab 3 \tab 9 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2164 \tab 52 \tab 2 \tab 7 \tab 8 \tab 64 \tab 50 \tab 11 \tab 71 \tab 4 \tab 105 \tab 12 \tab - \tab 24 \tab 11 \tab 7 \tab 2 \tab 19 \tab 2 \tab 1 \tab 9 \tab 61 \tab 10 \tab 15 \tab 4 \tab 2 \tab 6 \tab - \tab 3 \tab 2 \tab 8 \tab 3 \tab 1 \tab 3 \tab 3 \tab 7 \tab 5 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2266 \tab 61 \tab 2 \tab 8 \tab 5 \tab 62 \tab 38 \tab 9 \tab 72 \tab - \tab 84 \tab 13 \tab 2 \tab 18 \tab 11 \tab 5 \tab 5 \tab 9 \tab 3 \tab 3 \tab 9 \tab 61 \tab 5 \tab 11 \tab 5 \tab 4 \tab 2 \tab 1 \tab 2 \tab - \tab 4 \tab 2 \tab 2 \tab 2 \tab 3 \tab 1 \tab 4 \tab - \tab - \tab - \tab 9 \tab 2 \tab 2 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 3633 \tab 103 \tab 2 \tab 11 \tab 8 \tab 141 \tab 48 \tab 5 \tab 119 \tab 8 \tab 112 \tab 13 \tab 8 \tab 30 \tab 14 \tab 11 \tab 2 \tab 28 \tab 3 \tab 3 \tab 5 \tab 68 \tab 13 \tab 22 \tab 2 \tab 4 \tab 3 \tab 1 \tab 2 \tab 3 \tab 10 \tab 7 \tab - \tab 5 \tab 4 \tab 9 \tab 7 \tab - \tab - \tab - \tab 29 \tab 3 \tab - \tab 9 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1628 \tab 34 \tab 2 \tab 6 \tab 6 \tab 53 \tab 27 \tab 7 \tab 46 \tab 4 \tab 61 \tab 11 \tab 3 \tab 16 \tab 9 \tab 7 \tab - \tab 12 \tab - \tab 2 \tab 5 \tab 36 \tab 4 \tab 11 \tab 5 \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1616 \tab 45 \tab 4 \tab 9 \tab 3 \tab 58 \tab 20 \tab 7 \tab 54 \tab 4 \tab 63 \tab 8 \tab 2 \tab 13 \tab 12 \tab 6 \tab 4 \tab 7 \tab 3 \tab 2 \tab 6 \tab 40 \tab 9 \tab 12 \tab 3 \tab 2 \tab - \tab 2 \tab 3 \tab 3 \tab 5 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1613 \tab 40 \tab - \tab 5 \tab 2 \tab 59 \tab 19 \tab 7 \tab 43 \tab 5 \tab 50 \tab 14 \tab 4 \tab 7 \tab 5 \tab 6 \tab 3 \tab 13 \tab - \tab 1 \tab 5 \tab 26 \tab 6 \tab 13 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 3 \tab 5 \tab - \tab 3 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2052 \tab 36 \tab 1 \tab 9 \tab 5 \tab 76 \tab 17 \tab 5 \tab 64 \tab 4 \tab 68 \tab 12 \tab 5 \tab 21 \tab 9 \tab 9 \tab 2 \tab 17 \tab 4 \tab 1 \tab 4 \tab 43 \tab 11 \tab 8 \tab 4 \tab 2 \tab 1 \tab 2 \tab 5 \tab 1 \tab 8 \tab 4 \tab 1 \tab 2 \tab 2 \tab 7 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab 5 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2294 \tab 65 \tab 3 \tab 10 \tab 6 \tab 84 \tab 31 \tab 8 \tab 71 \tab 8 \tab 80 \tab 18 \tab 4 \tab 10 \tab 6 \tab 9 \tab - \tab 20 \tab 1 \tab 5 \tab 7 \tab 43 \tab 7 \tab 12 \tab 8 \tab 1 \tab 5 \tab 2 \tab 4 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 3 \tab 9 \tab 6 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 8 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1919 \tab 33 \tab 1 \tab 7 \tab 7 \tab 58 \tab 15 \tab 2 \tab 49 \tab 4 \tab 71 \tab 12 \tab 5 \tab 9 \tab 8 \tab 15 \tab 2 \tab 20 \tab 3 \tab 3 \tab 6 \tab 41 \tab 3 \tab 9 \tab 3 \tab - \tab 4 \tab 2 \tab 3 \tab - \tab 4 \tab - \tab 2 \tab - \tab 6 \tab 2 \tab 9 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 5 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab 3 \tab - \tab 5 \tab 2 \tab 17 \tab 13 \tab - \tab 22 \tab 1 \tab 25 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab 6 \tab - \tab 1 \tab 2 \tab 10 \tab 1 \tab 7 \tab - \tab 2 \tab 3 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 3 \tab - \tab 3389 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3544 \cr
   Total \tab 55271 \tab 784 \tab 33 \tab 120 \tab 108 \tab 1229 \tab 571 \tab 104 \tab 1162 \tab 69 \tab 1332 \tab 200 \tab 54 \tab 309 \tab 937 \tab 147 \tab 49 \tab 248 \tab 24 \tab 28 \tab 107 \tab 807 \tab 111 \tab 199 \tab 57 \tab 23 \tab 36 \tab 22 \tab 45 \tab 14 \tab 76 \tab 38 \tab 21 \tab 25 \tab 34 \tab 67 \tab 75 \tab 2 \tab 12 \tab 5 \tab 303 \tab 24 \tab 14 \tab 50 \tab 11 \tab 7421 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} africa (general)
\item \code{2} austria
\item \code{3} french canada
\item \code{4} other canada
\item \code{5} china
\item \code{6} czechoslovakia
\item \code{7} denmark
\item \code{8} england & wales
\item \code{9} finland
\item \code{10} france
\item \code{11} germany
\item \code{12} greece
\item \code{13} hungary
\item \code{14} ireland
\item \code{15} italy
\item \code{16} japan
\item \code{17} mexico
\item \code{18} netherlands
\item \code{19} norway
\item \code{20} philippines
\item \code{21} poland
\item \code{22} puerto rico
\item \code{23} russia
\item \code{24} scotland
\item \code{25} spain
\item \code{26} sweden
\item \code{27} switzerland
\item \code{28} west indies
\item \code{29} other
\item \code{30} american indian
\item \code{31} india
\item \code{32} portugal
\item \code{33} lithuania
\item \code{34} yugoslavia
\item \code{35} romania
\item \code{36} belgium
\item \code{37} arabic
\item \code{38} other spanish
\item \code{39} non-span west indies
\item \code{40} other asian
\item \code{41} other european
\item \code{97} american only
\item \code{101} turkey
\item \code{202} algeria
\item \code{203} congo
\item \code{204} egypt
\item \code{205} ethiopia
\item \code{206} kenya
\item \code{207} nigeria
\item \code{208} south africa
\item \code{209} sudan
\item \code{210} tanzania
\item \code{211} uganda
\item \code{299} other africa
\item \code{301} south korea
\item \code{302} bangladesh
\item \code{303} indonesia
\item \code{304} pakistan
\item \code{306} thailand
\item \code{307} vietnam
\item \code{401} iran
\item \code{402} iraq
\item \code{403} israel
\item \code{404} jordan
\item \code{405} saudi arabia
\item \code{406} syria
\item \code{407} uae
\item \code{408} yemen
\item \code{499} other middle east
\item \code{501} argentina
\item \code{502} bolivia
\item \code{503} brazil
\item \code{504} chile
\item \code{505} colombia
\item \code{506} ecuador
\item \code{507} french guiana
\item \code{508} guyana
\item \code{509} paraguay
\item \code{510} peru
\item \code{511} suriname
\item \code{512} uruguay
\item \code{513} venezuela
\item \code{599} other south america
\item \code{601} belize
\item \code{602} costa rica
\item \code{603} el salvador
\item \code{604} guatemala
\item \code{605} honduras
\item \code{606} nicaragua
\item \code{607} panama
\item \code{699} other central america
\item \code{799} other north america
\item \code{801} cuba
\item \code{802} haiti
\item \code{803} dominican republic
\item \code{804} jamaica
\item \code{899} other caribbean
\item \code{901} australia
\item \code{902} papua new guinea
\item \code{903} new zealand
\item \code{904} samoa
\item \code{999} other oceania
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
