% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_573.R
\name{ethnic}
\alias{ethnic}
\title{ethnic}
\source{
General Social Survey https://gss.norc.org
}
\description{
Country of family origin
}
\details{
Question 1596. From what countries or part of the world did your ancestors come?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab africa (general) \tab american indian \tab american only \tab arabic \tab austria \tab belgium \tab china \tab czechoslovakia \tab denmark \tab england & wales \tab finland \tab france \tab french canada \tab germany \tab greece \tab hungary \tab india \tab ireland \tab italy \tab lithuania \tab mexico \tab netherlands \tab no answer \tab non-span west indies \tab norway \tab other \tab other canada \tab other spanish \tab philippines \tab poland \tab portugal \tab puerto rico \tab romania \tab russia \tab scotland \tab spain \tab sweden \tab switzerland \tab uncodeable \tab west indies \tab yugoslavia \tab japan \tab other asian \tab other european \tab algeria \tab argentina \tab australia \tab bangladesh \tab belize \tab brazil \tab chile \tab colombia \tab congo \tab costa rica \tab cuba \tab dominican republic \tab ecuador \tab egypt \tab el salvador \tab ethiopia \tab guatemala \tab guyana \tab haiti \tab honduras \tab iap \tab iran \tab iraq \tab israel \tab jamaica \tab jordan \tab kenya \tab new zealand \tab nicaragua \tab nigeria \tab other africa \tab other caribbean \tab other central america \tab other middle east \tab other north america \tab other oceania \tab other south america \tab pakistan \tab panama \tab paraguay \tab peru \tab samoa \tab saudi arabia \tab south africa \tab south korea \tab suriname \tab syria \tab thailand \tab turkey \tab venezuela \tab vietnam \tab yemen \tab bolivia \tab french guiana \tab indonesia \tab skipped on web \tab sudan \tab tanzania \tab uganda \tab Total \cr
   1972 \tab 137 \tab 8 \tab 29 \tab 1 \tab 14 \tab 3 \tab 1 \tab 6 \tab 15 \tab 181 \tab 10 \tab 29 \tab 22 \tab 223 \tab 2 \tab 10 \tab 1 \tab 160 \tab 77 \tab 4 \tab 29 \tab 27 \tab 27 \tab 1 \tab 42 \tab 10 \tab 16 \tab 3 \tab 1 \tab 54 \tab 2 \tab 6 \tab 3 \tab 30 \tab 33 \tab 6 \tab 18 \tab 8 \tab 354 \tab 6 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 64 \tab 20 \tab 91 \tab - \tab 11 \tab 2 \tab 1 \tab 22 \tab 15 \tab 166 \tab 13 \tab 25 \tab 11 \tab 225 \tab 3 \tab 12 \tab - \tab 140 \tab 84 \tab 4 \tab 29 \tab 26 \tab 5 \tab 3 \tab 24 \tab 8 \tab 5 \tab 8 \tab 1 \tab 39 \tab 7 \tab 7 \tab 4 \tab 26 \tab 40 \tab 9 \tab 21 \tab 9 \tab 311 \tab 8 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 92 \tab 41 \tab 30 \tab 3 \tab 7 \tab 1 \tab - \tab 20 \tab 10 \tab 191 \tab 13 \tab 36 \tab 12 \tab 249 \tab 3 \tab 10 \tab 2 \tab 129 \tab 64 \tab 2 \tab 40 \tab 19 \tab 3 \tab 4 \tab 21 \tab 6 \tab 9 \tab 1 \tab 3 \tab 45 \tab 1 \tab 12 \tab 1 \tab 29 \tab 37 \tab 6 \tab 18 \tab 7 \tab 298 \tab 3 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 71 \tab 44 \tab 44 \tab - \tab 9 \tab 4 \tab - \tab 24 \tab 12 \tab 179 \tab 11 \tab 31 \tab 8 \tab 244 \tab 3 \tab 10 \tab 3 \tab 169 \tab 59 \tab 6 \tab 24 \tab 27 \tab 1 \tab - \tab 24 \tab 2 \tab 7 \tab 4 \tab 1 \tab 32 \tab 3 \tab 13 \tab 4 \tab 22 \tab 40 \tab 9 \tab 31 \tab 5 \tab 303 \tab 3 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 57 \tab 46 \tab 38 \tab 3 \tab 11 \tab 2 \tab 2 \tab 16 \tab 10 \tab 189 \tab 6 \tab 27 \tab 15 \tab 225 \tab 6 \tab 14 \tab 3 \tab 138 \tab 93 \tab 10 \tab 24 \tab 16 \tab 4 \tab 3 \tab 25 \tab 6 \tab 13 \tab 3 \tab 6 \tab 44 \tab 3 \tab 10 \tab 4 \tab 25 \tab 43 \tab 6 \tab 24 \tab 4 \tab 315 \tab 4 \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 99 \tab 44 \tab 26 \tab 1 \tab 15 \tab 2 \tab 4 \tab 18 \tab 12 \tab 166 \tab 12 \tab 26 \tab 13 \tab 253 \tab 3 \tab 12 \tab 2 \tab 116 \tab 69 \tab 5 \tab 23 \tab 27 \tab 9 \tab 3 \tab 22 \tab 10 \tab 15 \tab 4 \tab 4 \tab 34 \tab 6 \tab 9 \tab 2 \tab 23 \tab 44 \tab 9 \tab 20 \tab 7 \tab 345 \tab 5 \tab 7 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 68 \tab 38 \tab 48 \tab - \tab 6 \tab 2 \tab 3 \tab 24 \tab 8 \tab 170 \tab 10 \tab 31 \tab 17 \tab 249 \tab 1 \tab 10 \tab 2 \tab 162 \tab 64 \tab 4 \tab 16 \tab 20 \tab 11 \tab 3 \tab 17 \tab 12 \tab 5 \tab 5 \tab 6 \tab 46 \tab 5 \tab 8 \tab - \tab 10 \tab 37 \tab 6 \tab 25 \tab 7 \tab 369 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 64 \tab 47 \tab 31 \tab 4 \tab 14 \tab 1 \tab 1 \tab 19 \tab 11 \tab 183 \tab 14 \tab 23 \tab 8 \tab 229 \tab 8 \tab 10 \tab 3 \tab 140 \tab 66 \tab 10 \tab 23 \tab 18 \tab 17 \tab 2 \tab 21 \tab 8 \tab 10 \tab 11 \tab 2 \tab 46 \tab 5 \tab 16 \tab - \tab 14 \tab 41 \tab 9 \tab 28 \tab 2 \tab 303 \tab - \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 277 \tab 73 \tab 64 \tab 2 \tab 13 \tab 5 \tab 3 \tab 13 \tab 9 \tab 177 \tab 13 \tab 28 \tab 11 \tab 251 \tab 6 \tab 12 \tab 6 \tab 162 \tab 75 \tab 7 \tab 32 \tab 21 \tab 15 \tab 6 \tab 20 \tab 24 \tab 12 \tab 3 \tab 2 \tab 46 \tab 3 \tab 7 \tab 2 \tab 27 \tab 41 \tab 10 \tab 31 \tab 4 \tab 330 \tab 5 \tab 9 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 62 \tab 33 \tab 28 \tab 1 \tab 15 \tab 3 \tab 7 \tab 29 \tab 16 \tab 205 \tab 19 \tab 22 \tab 20 \tab 227 \tab 2 \tab 8 \tab 2 \tab 135 \tab 81 \tab 7 \tab 36 \tab 26 \tab 4 \tab 4 \tab 32 \tab 10 \tab 12 \tab 6 \tab 4 \tab 55 \tab 1 \tab 16 \tab 2 \tab 18 \tab 38 \tab 8 \tab 19 \tab 7 \tab 355 \tab 4 \tab 17 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 91 \tab 51 \tab 15 \tab - \tab 6 \tab 1 \tab 6 \tab 13 \tab 7 \tab 169 \tab 5 \tab 24 \tab 23 \tab 204 \tab 7 \tab 11 \tab 3 \tab 160 \tab 68 \tab 3 \tab 26 \tab 27 \tab 3 \tab 8 \tab 31 \tab 8 \tab 5 \tab 8 \tab 4 \tab 41 \tab 1 \tab 14 \tab 1 \tab 14 \tab 39 \tab 16 \tab 10 \tab 11 \tab 326 \tab 1 \tab 7 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 69 \tab 51 \tab 15 \tab 4 \tab 7 \tab 4 \tab 4 \tab 21 \tab 9 \tab 198 \tab 4 \tab 35 \tab 27 \tab 196 \tab 3 \tab 3 \tab 1 \tab 160 \tab 77 \tab 1 \tab 36 \tab 25 \tab 5 \tab 2 \tab 35 \tab 19 \tab 9 \tab 8 \tab 2 \tab 34 \tab 1 \tab 9 \tab 4 \tab 15 \tab 40 \tab 19 \tab 32 \tab 5 \tab 333 \tab 1 \tab 5 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 92 \tab 43 \tab 17 \tab 2 \tab 9 \tab - \tab 3 \tab 15 \tab 9 \tab 198 \tab 3 \tab 23 \tab 15 \tab 198 \tab 5 \tab 8 \tab 5 \tab 167 \tab 75 \tab 2 \tab 38 \tab 20 \tab 2 \tab 7 \tab 32 \tab 12 \tab 12 \tab 4 \tab 1 \tab 54 \tab 4 \tab 14 \tab 1 \tab 20 \tab 25 \tab 12 \tab 19 \tab 7 \tab 289 \tab 2 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 282 \tab 86 \tab 38 \tab 4 \tab 7 \tab 1 \tab 7 \tab 11 \tab 6 \tab 157 \tab 4 \tab 27 \tab 22 \tab 232 \tab 3 \tab 4 \tab 7 \tab 143 \tab 60 \tab 4 \tab 37 \tab 21 \tab 10 \tab 7 \tab 19 \tab 18 \tab 5 \tab 8 \tab 1 \tab 37 \tab 3 \tab 4 \tab 3 \tab 14 \tab 39 \tab 12 \tab 25 \tab 6 \tab 434 \tab 3 \tab 5 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 84 \tab 56 \tab 28 \tab 3 \tab 6 \tab 1 \tab 6 \tab 17 \tab 13 \tab 161 \tab - \tab 27 \tab 12 \tab 211 \tab 5 \tab 7 \tab 7 \tab 159 \tab 54 \tab 4 \tab 43 \tab 19 \tab 14 \tab 2 \tab 20 \tab 17 \tab 9 \tab 6 \tab 1 \tab 35 \tab 1 \tab 15 \tab 1 \tab 13 \tab 41 \tab 19 \tab 30 \tab 4 \tab 320 \tab - \tab 6 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 85 \tab 55 \tab 13 \tab 1 \tab 7 \tab 2 \tab 9 \tab 16 \tab 11 \tab 197 \tab - \tab 25 \tab 18 \tab 243 \tab 4 \tab 8 \tab 4 \tab 137 \tab 75 \tab 3 \tab 38 \tab 22 \tab 13 \tab 3 \tab 33 \tab 15 \tab 8 \tab 11 \tab 6 \tab 38 \tab 2 \tab 9 \tab 1 \tab 19 \tab 42 \tab 11 \tab 14 \tab 7 \tab 320 \tab 3 \tab 2 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 101 \tab 63 \tab 9 \tab 3 \tab 4 \tab 2 \tab 4 \tab 17 \tab 15 \tab 166 \tab 4 \tab 15 \tab 17 \tab 194 \tab 6 \tab 4 \tab 2 \tab 116 \tab 53 \tab 7 \tab 29 \tab 20 \tab 14 \tab 7 \tab 22 \tab 10 \tab 9 \tab 9 \tab 1 \tab 37 \tab 1 \tab 12 \tab 2 \tab 19 \tab 37 \tab 10 \tab 23 \tab 5 \tab 295 \tab - \tab 6 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 117 \tab 49 \tab 14 \tab 1 \tab 8 \tab 3 \tab 3 \tab 12 \tab 12 \tab 201 \tab 2 \tab 31 \tab 25 \tab 170 \tab 4 \tab 8 \tab 1 \tab 146 \tab 70 \tab 1 \tab 46 \tab 19 \tab 20 \tab 6 \tab 27 \tab 27 \tab 9 \tab 4 \tab 4 \tab 40 \tab 2 \tab 12 \tab - \tab 20 \tab 43 \tab 13 \tab 17 \tab 11 \tab 303 \tab - \tab 11 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 104 \tab 71 \tab 23 \tab 2 \tab 4 \tab 1 \tab 6 \tab 10 \tab 9 \tab 167 \tab 5 \tab 23 \tab 13 \tab 230 \tab 6 \tab 10 \tab 6 \tab 168 \tab 59 \tab 3 \tab 28 \tab 34 \tab 27 \tab 11 \tab 21 \tab 19 \tab 7 \tab 12 \tab 11 \tab 39 \tab 3 \tab 11 \tab 2 \tab 25 \tab 50 \tab 17 \tab 24 \tab 6 \tab 331 \tab 2 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 219 \tab 90 \tab 74 \tab 8 \tab 18 \tab 3 \tab 8 \tab 37 \tab 13 \tab 315 \tab 6 \tab 42 \tab 43 \tab 479 \tab 15 \tab 7 \tab 13 \tab 288 \tab 116 \tab 9 \tab 74 \tab 36 \tab 76 \tab 17 \tab 43 \tab 9 \tab 19 \tab 23 \tab 14 \tab 56 \tab 5 \tab 21 \tab 6 \tab 37 \tab 91 \tab 28 \tab 33 \tab 6 \tab 557 \tab 4 \tab 6 \tab 7 \tab 17 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 186 \tab 117 \tab 65 \tab 4 \tab 11 \tab 2 \tab 13 \tab 29 \tab 12 \tab 278 \tab 9 \tab 54 \tab 23 \tab 425 \tab 12 \tab 17 \tab 11 \tab 256 \tab 136 \tab 4 \tab 71 \tab 32 \tab 71 \tab 12 \tab 36 \tab 23 \tab 14 \tab 25 \tab 10 \tab 51 \tab 8 \tab 23 \tab 6 \tab 31 \tab 90 \tab 34 \tab 34 \tab 9 \tab 623 \tab - \tab 11 \tab 4 \tab 13 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 214 \tab 110 \tab 72 \tab 3 \tab 5 \tab 5 \tab 23 \tab 25 \tab 12 \tab 263 \tab 3 \tab 56 \tab 44 \tab 406 \tab 9 \tab 15 \tab 10 \tab 269 \tab 134 \tab 5 \tab 74 \tab 34 \tab 16 \tab 19 \tab 34 \tab 28 \tab 13 \tab 32 \tab 19 \tab 65 \tab 9 \tab 35 \tab 3 \tab 19 \tab 69 \tab 16 \tab 34 \tab 5 \tab 578 \tab 1 \tab 14 \tab 5 \tab 13 \tab 14 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 232 \tab 109 \tab 56 \tab 1 \tab 13 \tab 2 \tab 8 \tab 25 \tab 17 \tab 271 \tab 7 \tab 53 \tab 37 \tab 383 \tab 16 \tab 6 \tab 15 \tab 258 \tab 115 \tab 6 \tab 90 \tab 32 \tab 90 \tab 13 \tab 49 \tab 29 \tab 12 \tab 29 \tab 25 \tab 63 \tab 8 \tab 37 \tab - \tab 45 \tab 80 \tab 17 \tab 33 \tab 8 \tab 482 \tab 6 \tab 9 \tab 7 \tab 15 \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 228 \tab 118 \tab 65 \tab 8 \tab 10 \tab 6 \tab 14 \tab 19 \tab 14 \tab 231 \tab 7 \tab 48 \tab 32 \tab 330 \tab 11 \tab 12 \tab 11 \tab 266 \tab 125 \tab 5 \tab 84 \tab 27 \tab 39 \tab 16 \tab 40 \tab 6 \tab 16 \tab 34 \tab 18 \tab 67 \tab 6 \tab 31 \tab 6 \tab 39 \tab 82 \tab 18 \tab 35 \tab 4 \tab 591 \tab 4 \tab 3 \tab 7 \tab 15 \tab 17 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 221 \tab 120 \tab 56 \tab 13 \tab 8 \tab 3 \tab 27 \tab 20 \tab 15 \tab 286 \tab 11 \tab 41 \tab 10 \tab 341 \tab 11 \tab 10 \tab 21 \tab 283 \tab 119 \tab 5 \tab 130 \tab 28 \tab 12 \tab 13 \tab 34 \tab 10 \tab 10 \tab 28 \tab 22 \tab 42 \tab 11 \tab 27 \tab 5 \tab 31 \tab 71 \tab 24 \tab 29 \tab 9 \tab 599 \tab 1 \tab 4 \tab 10 \tab 24 \tab 17 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 389 \tab 160 \tab 130 \tab 12 \tab 7 \tab 10 \tab 24 \tab 57 \tab 22 \tab 389 \tab 11 \tab 57 \tab 19 \tab 482 \tab 21 \tab 14 \tab 25 \tab 432 \tab 196 \tab 5 \tab 333 \tab 47 \tab 14 \tab 26 \tab 44 \tab 20 \tab 19 \tab 107 \tab 25 \tab 88 \tab 16 \tab 53 \tab 6 \tab 44 \tab 119 \tab 51 \tab 52 \tab 21 \tab 878 \tab 4 \tab 12 \tab 19 \tab 29 \tab 21 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 150 \tab 78 \tab 44 \tab 5 \tab 4 \tab 6 \tab 13 \tab 10 \tab 9 \tab 173 \tab 4 \tab 31 \tab 10 \tab 246 \tab 6 \tab 5 \tab 23 \tab 218 \tab 92 \tab 4 \tab 129 \tab 23 \tab 8 \tab 22 \tab 27 \tab 9 \tab 10 \tab 41 \tab 13 \tab 27 \tab 10 \tab 19 \tab 5 \tab 21 \tab 56 \tab 30 \tab 27 \tab 8 \tab 374 \tab 3 \tab 4 \tab 4 \tab 15 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 152 \tab 77 \tab 60 \tab 4 \tab 5 \tab 8 \tab 11 \tab 12 \tab 15 \tab 162 \tab 6 \tab 27 \tab 14 \tab 242 \tab 9 \tab 5 \tab 14 \tab 198 \tab 82 \tab 4 \tab 127 \tab 21 \tab 8 \tab 17 \tab 23 \tab 10 \tab 7 \tab 44 \tab 8 \tab 44 \tab 15 \tab 23 \tab 4 \tab 18 \tab 54 \tab 15 \tab 18 \tab 9 \tab 418 \tab 3 \tab 5 \tab 13 \tab 23 \tab 10 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 163 \tab 83 \tab 61 \tab 3 \tab 8 \tab 1 \tab 16 \tab 6 \tab 7 \tab 160 \tab 9 \tab 25 \tab 17 \tab 236 \tab 10 \tab 6 \tab 23 \tab 196 \tab 89 \tab 2 \tab 134 \tab 18 \tab 7 \tab 13 \tab 16 \tab 44 \tab 6 \tab 36 \tab 12 \tab 38 \tab 9 \tab 18 \tab 4 \tab 14 \tab 37 \tab 37 \tab 23 \tab 8 \tab 343 \tab 2 \tab 8 \tab 3 \tab 17 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 183 \tab 100 \tab 50 \tab 9 \tab 11 \tab 1 \tab 22 \tab 16 \tab 7 \tab 216 \tab 8 \tab 28 \tab 22 \tab 282 \tab 11 \tab 8 \tab 31 \tab 245 \tab 128 \tab 4 \tab 188 \tab 17 \tab 8 \tab 28 \tab 20 \tab 9 \tab 12 \tab 87 \tab 15 \tab 43 \tab 11 \tab 30 \tab 2 \tab 13 \tab 56 \tab 36 \tab 27 \tab 3 \tab 500 \tab 3 \tab 4 \tab 7 \tab 18 \tab 19 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 242 \tab 143 \tab 61 \tab 8 \tab 4 \tab 6 \tab 21 \tab 21 \tab 9 \tab 210 \tab 9 \tab 34 \tab 23 \tab 311 \tab 15 \tab 9 \tab 22 \tab 244 \tab 143 \tab 7 \tab 187 \tab 23 \tab 6 \tab 18 \tab 29 \tab 12 \tab 22 \tab 69 \tab 22 \tab 71 \tab 14 \tab 35 \tab 5 \tab 33 \tab 68 \tab 26 \tab 34 \tab 14 \tab 574 \tab 6 \tab 10 \tab 9 \tab 24 \tab 14 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 182 \tab 84 \tab 65 \tab 9 \tab 11 \tab 1 \tab 17 \tab 13 \tab 11 \tab 174 \tab 11 \tab 27 \tab 24 \tab 239 \tab 8 \tab 7 \tab 17 \tab 207 \tab 92 \tab 6 \tab 181 \tab 11 \tab 9 \tab 23 \tab 21 \tab 12 \tab 17 \tab 56 \tab 18 \tab 38 \tab 5 \tab 31 \tab 4 \tab 23 \tab 60 \tab 34 \tab 20 \tab 9 \tab 506 \tab 1 \tab 4 \tab 6 \tab 23 \tab 31 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 122 \tab - \tab - \tab 21 \tab 11 \tab 37 \tab 50 \tab 47 \tab 599 \tab 15 \tab 91 \tab 13 \tab 507 \tab 21 \tab 22 \tab 39 \tab 380 \tab 186 \tab 14 \tab 179 \tab 34 \tab - \tab - \tab 44 \tab - \tab 17 \tab - \tab 16 \tab 90 \tab 16 \tab 37 \tab 4 \tab 36 \tab 121 \tab 45 \tab 41 \tab 18 \tab - \tab - \tab 8 \tab 18 \tab 21 \tab 129 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 9 \tab 1 \tab 9 \tab 5 \tab 2 \tab 18 \tab 18 \tab 9 \tab 5 \tab 7 \tab 7 \tab 5 \tab 2 \tab 4 \tab 4 \tab 438 \tab 3 \tab 4 \tab 5 \tab 20 \tab 2 \tab 1 \tab 1 \tab 3 \tab 22 \tab 154 \tab 7 \tab 9 \tab 8 \tab 91 \tab 5 \tab 20 \tab 3 \tab 5 \tab 1 \tab 4 \tab 2 \tab 1 \tab 16 \tab 24 \tab 1 \tab 3 \tab 6 \tab 1 \tab 4 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab 173 \tab - \tab - \tab 14 \tab 7 \tab 24 \tab 28 \tab 30 \tab 399 \tab 19 \tab 67 \tab 12 \tab 379 \tab 9 \tab 9 \tab 31 \tab 300 \tab 170 \tab 9 \tab 246 \tab 43 \tab 1 \tab - \tab 53 \tab - \tab 17 \tab - \tab 30 \tab 107 \tab 18 \tab 43 \tab 4 \tab 22 \tab 107 \tab 51 \tab 45 \tab 8 \tab - \tab - \tab 10 \tab 10 \tab 7 \tab 95 \tab 1 \tab 2 \tab - \tab 5 \tab 1 \tab 4 \tab 1 \tab 15 \tab 8 \tab 3 \tab 18 \tab 13 \tab 5 \tab 12 \tab 15 \tab 9 \tab 19 \tab - \tab 9 \tab 6 \tab 238 \tab 2 \tab 1 \tab 5 \tab 29 \tab 1 \tab 9 \tab - \tab 2 \tab 49 \tab 199 \tab 21 \tab 10 \tab 12 \tab 108 \tab 1 \tab 8 \tab 6 \tab 7 \tab - \tab 4 \tab 1 \tab 2 \tab 37 \tab 11 \tab - \tab 3 \tab 3 \tab 1 \tab 5 \tab 10 \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 4 \tab 3544 \cr
   Total \tab 4777 \tab 2603 \tab 1460 \tab 122 \tab 323 \tab 112 \tab 348 \tab 691 \tab 449 \tab 7547 \tab 283 \tab 1189 \tab 652 \tab 9541 \tab 264 \tab 323 \tab 366 \tab 6847 \tab 3246 \tab 176 \tab 2824 \tab 860 \tab 569 \tab 319 \tab 1001 \tab 462 \tab 388 \tab 729 \tab 328 \tab 1686 \tab 215 \tab 667 \tab 101 \tab 809 \tab 1911 \tab 669 \tab 914 \tab 261 \tab 13257 \tab 89 \tab 215 \tab 192 \tab 274 \tab 401 \tab 2 \tab 4 \tab 1 \tab 7 \tab 2 \tab 13 \tab 2 \tab 24 \tab 13 \tab 5 \tab 36 \tab 31 \tab 14 \tab 17 \tab 22 \tab 16 \tab 24 \tab 2 \tab 13 \tab 10 \tab 676 \tab 5 \tab 5 \tab 10 \tab 49 \tab 3 \tab 10 \tab 1 \tab 5 \tab 71 \tab 353 \tab 28 \tab 19 \tab 20 \tab 199 \tab 6 \tab 28 \tab 9 \tab 12 \tab 1 \tab 8 \tab 3 \tab 3 \tab 53 \tab 35 \tab 1 \tab 6 \tab 9 \tab 2 \tab 9 \tab 16 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 4 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} africa (general)
\item \code{2} austria
\item \code{3} french canada
\item \code{4} other canada
\item \code{5} china
\item \code{6} czechoslovakia
\item \code{7} denmark
\item \code{8} england & wales
\item \code{9} finland
\item \code{10} france
\item \code{11} germany
\item \code{12} greece
\item \code{13} hungary
\item \code{14} ireland
\item \code{15} italy
\item \code{16} japan
\item \code{17} mexico
\item \code{18} netherlands
\item \code{19} norway
\item \code{20} philippines
\item \code{21} poland
\item \code{22} puerto rico
\item \code{23} russia
\item \code{24} scotland
\item \code{25} spain
\item \code{26} sweden
\item \code{27} switzerland
\item \code{28} west indies
\item \code{29} other
\item \code{30} american indian
\item \code{31} india
\item \code{32} portugal
\item \code{33} lithuania
\item \code{34} yugoslavia
\item \code{35} romania
\item \code{36} belgium
\item \code{37} arabic
\item \code{38} other spanish
\item \code{39} non-span west indies
\item \code{40} other asian
\item \code{41} other european
\item \code{97} american only
\item \code{101} turkey
\item \code{202} algeria
\item \code{203} congo
\item \code{204} egypt
\item \code{205} ethiopia
\item \code{206} kenya
\item \code{207} nigeria
\item \code{208} south africa
\item \code{209} sudan
\item \code{210} tanzania
\item \code{211} uganda
\item \code{299} other africa
\item \code{301} south korea
\item \code{302} bangladesh
\item \code{303} indonesia
\item \code{304} pakistan
\item \code{306} thailand
\item \code{307} vietnam
\item \code{401} iran
\item \code{402} iraq
\item \code{403} israel
\item \code{404} jordan
\item \code{405} saudi arabia
\item \code{406} syria
\item \code{407} uae
\item \code{408} yemen
\item \code{499} other middle east
\item \code{501} argentina
\item \code{502} bolivia
\item \code{503} brazil
\item \code{504} chile
\item \code{505} colombia
\item \code{506} ecuador
\item \code{507} french guiana
\item \code{508} guyana
\item \code{509} paraguay
\item \code{510} peru
\item \code{511} suriname
\item \code{512} uruguay
\item \code{513} venezuela
\item \code{599} other south america
\item \code{601} belize
\item \code{602} costa rica
\item \code{603} el salvador
\item \code{604} guatemala
\item \code{605} honduras
\item \code{606} nicaragua
\item \code{607} panama
\item \code{699} other central america
\item \code{799} other north america
\item \code{801} cuba
\item \code{802} haiti
\item \code{803} dominican republic
\item \code{804} jamaica
\item \code{899} other caribbean
\item \code{901} australia
\item \code{902} papua new guinea
\item \code{903} new zealand
\item \code{904} samoa
\item \code{999} other oceania
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
