% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_574.R
\name{ethnum}
\alias{ethnum}
\title{Type of response about ethnicity:r}
\source{
General Social Survey https://gss.norc.org
}
\description{
ethnum
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab can't name any \tab names one country \tab names two or more countries, can't choose \tab names two or more countries, chooses one \tab uncodeable \tab iap \tab Total \cr
   1972 \tab 165 \tab 936 \tab 189 \tab 296 \tab 27 \tab - \tab 1613 \cr
   1973 \tab 150 \tab 872 \tab 161 \tab 316 \tab 5 \tab - \tab 1504 \cr
   1974 \tab 137 \tab 819 \tab 161 \tab 364 \tab 3 \tab - \tab 1484 \cr
   1975 \tab 155 \tab 795 \tab 148 \tab 391 \tab 1 \tab - \tab 1490 \cr
   1976 \tab 152 \tab 804 \tab 163 \tab 376 \tab 4 \tab - \tab 1499 \cr
   1977 \tab 179 \tab 842 \tab 166 \tab 334 \tab 9 \tab - \tab 1530 \cr
   1978 \tab 185 \tab 789 \tab 184 \tab 363 \tab 11 \tab - \tab 1532 \cr
   1980 \tab 165 \tab 734 \tab 138 \tab 414 \tab 17 \tab - \tab 1468 \cr
   1982 \tab 203 \tab 1075 \tab 127 \tab 440 \tab 15 \tab - \tab 1860 \cr
   1983 \tab 183 \tab 760 \tab 172 \tab 480 \tab 4 \tab - \tab 1599 \cr
   1984 \tab 170 \tab 679 \tab 156 \tab 465 \tab 3 \tab - \tab 1473 \cr
   1985 \tab 162 \tab 654 \tab 171 \tab 542 \tab 5 \tab - \tab 1534 \cr
   1986 \tab 160 \tab 680 \tab 129 \tab 499 \tab 2 \tab - \tab 1470 \cr
   1987 \tab 254 \tab 878 \tab 180 \tab 497 \tab 10 \tab - \tab 1819 \cr
   1988 \tab 169 \tab 708 \tab 151 \tab 439 \tab 14 \tab - \tab 1481 \cr
   1989 \tab 125 \tab 766 \tab 195 \tab 438 \tab 13 \tab - \tab 1537 \cr
   1990 \tab 149 \tab 638 \tab 132 \tab 439 \tab 14 \tab - \tab 1372 \cr
   1991 \tab 152 \tab 748 \tab 138 \tab 459 \tab 20 \tab - \tab 1517 \cr
   1993 \tab 175 \tab 726 \tab 141 \tab 537 \tab 27 \tab - \tab 1606 \cr
   1994 \tab 270 \tab 1414 \tab 276 \tab 954 \tab 78 \tab - \tab 2992 \cr
   1996 \tab 363 \tab 1400 \tab 269 \tab 845 \tab 27 \tab - \tab 2904 \cr
   1998 \tab 282 \tab 1437 \tab 298 \tab 800 \tab 15 \tab - \tab 2832 \cr
   2000 \tab 261 \tab 1467 \tab 305 \tab 774 \tab 10 \tab - \tab 2817 \cr
   2002 \tab 208 \tab 1079 \tab 404 \tab 1064 \tab 10 \tab - \tab 2765 \cr
   2004 \tab 217 \tab 1351 \tab 380 \tab 850 \tab 14 \tab - \tab 2812 \cr
   2006 \tab 314 \tab 2307 \tab 564 \tab 1311 \tab 14 \tab - \tab 4510 \cr
   2008 \tab 160 \tab 1088 \tab 214 \tab 553 \tab 8 \tab - \tab 2023 \cr
   2010 \tab 185 \tab 985 \tab 230 \tab 637 \tab 7 \tab - \tab 2044 \cr
   2012 \tab 158 \tab 1098 \tab 186 \tab 527 \tab 5 \tab - \tab 1974 \cr
   2014 \tab 209 \tab 1286 \tab 294 \tab 741 \tab 8 \tab - \tab 2538 \cr
   2016 \tab 271 \tab 1306 \tab 303 \tab 981 \tab 6 \tab - \tab 2867 \cr
   2018 \tab 292 \tab 1077 \tab 216 \tab 755 \tab 8 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 6380 \tab 32198 \tab 6941 \tab 18881 \tab 414 \tab 7576 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} names one country
\item \code{2} names two or more countries, chooses one
\item \code{3} names two or more countries, can't choose
\item \code{4} can't name any
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
