% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_553.R
\name{evpaidsx}
\alias{evpaidsx}
\title{evpaidsx}
\source{
General Social Survey https://gss.norc.org
}
\description{
Ever have sex paid for or being paid since 18
}
\details{
Question 1554. Thinking about the time since your 18th birthday, have you ever had sex with a person you paid or who paid you for sex?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab don't know \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 221 \tab 1166 \tab 28 \tab 102 \tab - \tab 1517 \cr
   1993 \tab 114 \tab 1327 \tab 41 \tab 124 \tab - \tab 1606 \cr
   1994 \tab 201 \tab 2439 \tab 114 \tab 238 \tab - \tab 2992 \cr
   1996 \tab 247 \tab 2337 \tab 94 \tab 226 \tab - \tab 2904 \cr
   1998 \tab 381 \tab 2170 \tab 101 \tab 180 \tab - \tab 2832 \cr
   2000 \tab 417 \tab 2109 \tab 99 \tab 191 \tab 1 \tab 2817 \cr
   2002 \tab 489 \tab 2048 \tab 50 \tab 175 \tab 3 \tab 2765 \cr
   2004 \tab 584 \tab 2030 \tab 31 \tab 164 \tab 3 \tab 2812 \cr
   2006 \tab 2096 \tab 2189 \tab 25 \tab 198 \tab 2 \tab 4510 \cr
   2008 \tab 240 \tab 1600 \tab 21 \tab 160 \tab 2 \tab 2023 \cr
   2010 \tab 202 \tab 1675 \tab 32 \tab 129 \tab 6 \tab 2044 \cr
   2012 \tab 238 \tab 1586 \tab 31 \tab 114 \tab 5 \tab 1974 \cr
   2014 \tab 188 \tab 2152 \tab 57 \tab 139 \tab 2 \tab 2538 \cr
   2016 \tab 1096 \tab 1628 \tab 23 \tab 119 \tab 1 \tab 2867 \cr
   2018 \tab 942 \tab 1302 \tab 16 \tab 86 \tab 2 \tab 2348 \cr
   2021 \tab 1705 \tab 2132 \tab 31 \tab 143 \tab 21 \tab 4032 \cr
   2022 \tab 1898 \tab 1452 \tab 59 \tab 100 \tab 35 \tab 3544 \cr
   Total \tab 37524 \tab 31342 \tab 853 \tab 2588 \tab 83 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
