% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_553.R
\name{evstray}
\alias{evstray}
\title{evstray}
\source{
General Social Survey https://gss.norc.org
}
\description{
Have sex other than spouse while married
}
\details{
Question 1555. Have you ever had sex with someone other than your husband or wife while you were married?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab never married \tab no \tab no answer \tab yes \tab don't know \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 221 \tab 279 \tab 853 \tab 18 \tab 146 \tab - \tab 1517 \cr
   1993 \tab 114 \tab 258 \tab 982 \tab 57 \tab 195 \tab - \tab 1606 \cr
   1994 \tab 201 \tab 586 \tab 1781 \tab 70 \tab 354 \tab - \tab 2992 \cr
   1996 \tab 258 \tab 615 \tab 1618 \tab 43 \tab 369 \tab 1 \tab 2904 \cr
   1998 \tab 383 \tab 587 \tab 1481 \tab 57 \tab 324 \tab - \tab 2832 \cr
   2000 \tab 417 \tab 602 \tab 1418 \tab 71 \tab 308 \tab 1 \tab 2817 \cr
   2002 \tab 502 \tab 594 \tab 1307 \tab 49 \tab 311 \tab 2 \tab 2765 \cr
   2004 \tab 573 \tab 540 \tab 1360 \tab 45 \tab 291 \tab 3 \tab 2812 \cr
   2006 \tab 2096 \tab 623 \tab 1414 \tab 26 \tab 350 \tab 1 \tab 4510 \cr
   2008 \tab 240 \tab 491 \tab 1046 \tab 28 \tab 214 \tab 4 \tab 2023 \cr
   2010 \tab 202 \tab 523 \tab 1058 \tab 31 \tab 227 \tab 3 \tab 2044 \cr
   2012 \tab 238 \tab 486 \tab 1003 \tab 32 \tab 213 \tab 2 \tab 1974 \cr
   2014 \tab 188 \tab 636 \tab 1375 \tab 53 \tab 285 \tab 1 \tab 2538 \cr
   2016 \tab 1096 \tab 523 \tab 1003 \tab 28 \tab 217 \tab - \tab 2867 \cr
   2018 \tab 942 \tab 425 \tab 806 \tab 14 \tab 159 \tab 2 \tab 2348 \cr
   2021 \tab 1615 \tab 646 \tab 1408 \tab 49 \tab 297 \tab 17 \tab 4032 \cr
   2022 \tab 1300 \tab 1101 \tab 884 \tab 73 \tab 170 \tab 16 \tab 3544 \cr
   Total \tab 36851 \tab 9515 \tab 20797 \tab 744 \tab 4430 \tab 53 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} never married
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
