% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_368.R
\name{exptext}
\alias{exptext}
\title{Why is it better to test drug this way}
\source{
General Social Survey https://gss.norc.org
}
\description{
exptext
}
\details{
Question 1042. Why is it better to test the drug this way?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab correct - control group \tab correct - don't know \tab correct - vague reason \tab correct - wrong reason \tab no answer to closed ended question \tab uncodeable \tab wrong - don't know \tab wrong - other reason \tab wrong - reservations about control group \tab wrong - reservations about sample \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2646 \tab 300 \tab 35 \tab 450 \tab 682 \tab 84 \tab 2 \tab 21 \tab 106 \tab 27 \tab 157 \tab 4510 \cr
   2008 \tab 629 \tab 199 \tab 14 \tab 362 \tab 513 \tab 2 \tab 33 \tab 8 \tab 106 \tab 37 \tab 120 \tab 2023 \cr
   2010 \tab 1364 \tab 240 \tab 17 \tab 114 \tab 193 \tab 17 \tab - \tab 6 \tab 21 \tab 10 \tab 62 \tab 2044 \cr
   2012 \tab 1038 \tab 154 \tab 33 \tab 144 \tab 460 \tab 11 \tab - \tab 4 \tab 55 \tab 23 \tab 52 \tab 1974 \cr
   2014 \tab 1366 \tab 194 \tab 12 \tab 394 \tab 347 \tab 24 \tab - \tab 2 \tab 86 \tab 20 \tab 93 \tab 2538 \cr
   2016 \tab 1526 \tab 264 \tab 34 \tab 428 \tab 364 \tab 19 \tab - \tab 12 \tab 80 \tab 36 \tab 104 \tab 2867 \cr
   2018 \tab 1212 \tab 288 \tab 28 \tab 281 \tab 303 \tab 14 \tab - \tab 17 \tab 57 \tab 44 \tab 104 \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 63867 \tab 1639 \tab 173 \tab 2173 \tab 2862 \tab 171 \tab 35 \tab 70 \tab 511 \tab 197 \tab 692 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} correct - control group
\item \code{2} correct - vague reason
\item \code{3} correct - wrong reason
\item \code{4} correct - don't know
\item \code{5} wrong - reservations about control group
\item \code{6} wrong - reservations about sample
\item \code{7} wrong - other reason
\item \code{8} wrong - don't know
\item \code{9} no answer to closed ended question
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
