% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_50.R
\name{fair}
\alias{fair}
\title{People fair or try to take advantage}
\source{
General Social Survey https://gss.norc.org
}
\description{
fair
}
\details{
Question 162. Do you think most people would try to take advantage of you if they got a chance, or would they try to be fair?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab depends \tab don't know \tab no answer \tab would take advantage of you \tab would try to be fair \tab iap \tab not available in this year \tab Total \cr
   1972 \tab 96 \tab 19 \tab 2 \tab 542 \tab 954 \tab - \tab - \tab 1613 \cr
   1973 \tab 72 \tab 8 \tab 1 \tab 562 \tab 861 \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab - \tab - \tab - \tab - \tab 1484 \tab - \tab 1484 \cr
   1975 \tab 105 \tab 12 \tab 2 \tab 454 \tab 917 \tab - \tab - \tab 1490 \cr
   1976 \tab 62 \tab 8 \tab - \tab 542 \tab 887 \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab - \tab - \tab - \tab - \tab 1530 \tab - \tab 1530 \cr
   1978 \tab 78 \tab 10 \tab 7 \tab 454 \tab 983 \tab - \tab - \tab 1532 \cr
   1980 \tab 63 \tab 18 \tab - \tab 505 \tab 882 \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab - \tab - \tab - \tab - \tab 1860 \tab - \tab 1860 \cr
   1983 \tab 82 \tab 14 \tab - \tab 553 \tab 950 \tab - \tab - \tab 1599 \cr
   1984 \tab 47 \tab 3 \tab 3 \tab 507 \tab 913 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab - \tab - \tab - \tab - \tab 1534 \tab - \tab 1534 \cr
   1986 \tab 55 \tab 9 \tab 5 \tab 490 \tab 911 \tab - \tab - \tab 1470 \cr
   1987 \tab 75 \tab 18 \tab 9 \tab 751 \tab 966 \tab - \tab - \tab 1819 \cr
   1988 \tab 51 \tab 12 \tab 1 \tab 335 \tab 598 \tab 484 \tab - \tab 1481 \cr
   1989 \tab 54 \tab 6 \tab - \tab 376 \tab 599 \tab 502 \tab - \tab 1537 \cr
   1990 \tab 53 \tab 6 \tab - \tab 325 \tab 515 \tab 473 \tab - \tab 1372 \cr
   1991 \tab 59 \tab 7 \tab 6 \tab 366 \tab 579 \tab 500 \tab - \tab 1517 \cr
   1993 \tab 74 \tab 8 \tab 2 \tab 384 \tab 589 \tab 549 \tab - \tab 1606 \cr
   1994 \tab 134 \tab 16 \tab 9 \tab 787 \tab 1065 \tab 981 \tab - \tab 2992 \cr
   1996 \tab 132 \tab 22 \tab 13 \tab 804 \tab 954 \tab 979 \tab - \tab 2904 \cr
   1998 \tab 167 \tab 19 \tab 3 \tab 725 \tab 997 \tab 921 \tab - \tab 2832 \cr
   2000 \tab 172 \tab 24 \tab 9 \tab 718 \tab 973 \tab 921 \tab - \tab 2817 \cr
   2002 \tab 97 \tab 3 \tab - \tab 361 \tab 451 \tab 1853 \tab - \tab 2765 \cr
   2004 \tab 80 \tab 4 \tab 1 \tab 347 \tab 444 \tab 1936 \tab - \tab 2812 \cr
   2006 \tab 236 \tab 10 \tab - \tab 781 \tab 962 \tab 2521 \tab - \tab 4510 \cr
   2008 \tab 105 \tab 5 \tab 1 \tab 553 \tab 701 \tab 658 \tab - \tab 2023 \cr
   2010 \tab 121 \tab 10 \tab 2 \tab 524 \tab 720 \tab 667 \tab - \tab 2044 \cr
   2012 \tab 89 \tab 11 \tab 1 \tab 553 \tab 684 \tab 636 \tab - \tab 1974 \cr
   2014 \tab 137 \tab 6 \tab - \tab 675 \tab 869 \tab 851 \tab - \tab 2538 \cr
   2016 \tab 114 \tab 11 \tab 6 \tab 797 \tab 1028 \tab 911 \tab - \tab 2867 \cr
   2018 \tab 112 \tab 12 \tab 1 \tab 671 \tab 767 \tab 785 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 107 \tab 12 \tab 1 \tab 566 \tab 533 \tab 2325 \tab - \tab 3544 \cr
   Total \tab 2829 \tab 323 \tab 85 \tab 16008 \tab 23252 \tab 25861 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} would take advantage of you
\item \code{2} would try to be fair
\item \code{3} depends
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
