% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_14.R
\name{famdif16}
\alias{famdif16}
\title{Reason not living with parents}
\source{
General Social Survey https://gss.norc.org
}
\description{
famdif16
}
\details{
Question 27a. If not living with both own mother and father: What happened?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab father absent in armed forces \tab no answer \tab one or both parents died \tab one or both parents in institution \tab other \tab parents divorced or separated \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1174 \tab 1 \tab 1 \tab 199 \tab 2 \tab 21 \tab 106 \tab - \tab - \tab 1504 \cr
   1974 \tab 1126 \tab 1 \tab 3 \tab 208 \tab - \tab 25 \tab 121 \tab - \tab - \tab 1484 \cr
   1975 \tab 1140 \tab - \tab 1 \tab 183 \tab 2 \tab 24 \tab 137 \tab 3 \tab - \tab 1490 \cr
   1976 \tab 1139 \tab 1 \tab 9 \tab 215 \tab 2 \tab 11 \tab 121 \tab 1 \tab - \tab 1499 \cr
   1977 \tab 1133 \tab 1 \tab 5 \tab 215 \tab 4 \tab 25 \tab 144 \tab 3 \tab - \tab 1530 \cr
   1978 \tab 1157 \tab - \tab 2 \tab 202 \tab 2 \tab 34 \tab 132 \tab 3 \tab - \tab 1532 \cr
   1980 \tab 1070 \tab 1 \tab 2 \tab 217 \tab - \tab 46 \tab 131 \tab 1 \tab - \tab 1468 \cr
   1982 \tab 1334 \tab 1 \tab 6 \tab 230 \tab 2 \tab 55 \tab 226 \tab 6 \tab - \tab 1860 \cr
   1983 \tab 1225 \tab - \tab 3 \tab 187 \tab 1 \tab 24 \tab 157 \tab 2 \tab - \tab 1599 \cr
   1984 \tab 1100 \tab - \tab 24 \tab 169 \tab 2 \tab 36 \tab 137 \tab 5 \tab - \tab 1473 \cr
   1985 \tab 1160 \tab - \tab 1 \tab 173 \tab 1 \tab 46 \tab 144 \tab 9 \tab - \tab 1534 \cr
   1986 \tab 1082 \tab 1 \tab 20 \tab 165 \tab - \tab 41 \tab 160 \tab 1 \tab - \tab 1470 \cr
   1987 \tab 1316 \tab 3 \tab 28 \tab 195 \tab 1 \tab 73 \tab 197 \tab 6 \tab - \tab 1819 \cr
   1988 \tab 1072 \tab 1 \tab 14 \tab 157 \tab 1 \tab 44 \tab 183 \tab 9 \tab - \tab 1481 \cr
   1989 \tab 1153 \tab - \tab 10 \tab 139 \tab 2 \tab 57 \tab 173 \tab 3 \tab - \tab 1537 \cr
   1990 \tab 992 \tab 1 \tab 7 \tab 143 \tab 1 \tab 52 \tab 171 \tab 5 \tab - \tab 1372 \cr
   1991 \tab 1102 \tab 1 \tab 7 \tab 165 \tab 1 \tab 41 \tab 194 \tab 6 \tab - \tab 1517 \cr
   1993 \tab 1192 \tab - \tab - \tab 143 \tab - \tab 72 \tab 192 \tab 7 \tab - \tab 1606 \cr
   1994 \tab 2101 \tab 2 \tab 21 \tab 303 \tab 1 \tab 125 \tab 430 \tab 9 \tab - \tab 2992 \cr
   1996 \tab 2012 \tab - \tab 34 \tab 271 \tab 4 \tab 145 \tab 433 \tab 5 \tab - \tab 2904 \cr
   1998 \tab 1959 \tab 3 \tab 50 \tab 231 \tab 5 \tab 142 \tab 430 \tab 12 \tab - \tab 2832 \cr
   2000 \tab 1936 \tab 2 \tab 92 \tab 238 \tab 8 \tab 134 \tab 398 \tab 9 \tab - \tab 2817 \cr
   2002 \tab 1861 \tab 2 \tab 8 \tab 232 \tab 8 \tab 166 \tab 485 \tab 3 \tab - \tab 2765 \cr
   2004 \tab 1967 \tab 3 \tab 9 \tab 192 \tab 4 \tab 162 \tab 473 \tab 2 \tab - \tab 2812 \cr
   2006 \tab 3603 \tab 1 \tab 4 \tab 217 \tab 15 \tab 155 \tab 512 \tab 3 \tab - \tab 4510 \cr
   2008 \tab 1375 \tab 1 \tab 2 \tab 163 \tab 4 \tab 119 \tab 357 \tab 2 \tab - \tab 2023 \cr
   2010 \tab 1358 \tab 3 \tab 5 \tab 151 \tab 7 \tab 129 \tab 388 \tab 3 \tab - \tab 2044 \cr
   2012 \tab 1303 \tab 5 \tab 2 \tab 130 \tab 7 \tab 155 \tab 369 \tab 3 \tab - \tab 1974 \cr
   2014 \tab 1672 \tab 3 \tab 4 \tab 154 \tab 11 \tab 176 \tab 514 \tab 4 \tab - \tab 2538 \cr
   2016 \tab 1845 \tab 4 \tab 12 \tab 195 \tab 13 \tab 215 \tab 581 \tab 2 \tab - \tab 2867 \cr
   2018 \tab 1499 \tab 7 \tab 8 \tab 157 \tab 16 \tab 173 \tab 483 \tab 5 \tab - \tab 2348 \cr
   2021 \tab 3776 \tab 3 \tab - \tab 66 \tab 7 \tab 29 \tab 149 \tab - \tab 2 \tab 4032 \cr
   2022 \tab 2950 \tab 4 \tab 5 \tab 152 \tab 8 \tab 76 \tab 344 \tab 4 \tab 1 \tab 3544 \cr
   Total \tab 53497 \tab 56 \tab 399 \tab 6157 \tab 142 \tab 2828 \tab 9172 \tab 136 \tab 3 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} one or both parents died
\item \code{2} parents divorced or separated
\item \code{3} father absent in armed forces
\item \code{4} one or both parents in institution
\item \code{5} other
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
