% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{family16}
\alias{family16}
\title{family16}
\source{
General Social Survey https://gss.norc.org
}
\description{
Living with parents when 16 yrs old
}
\details{
Question 27. Were you living with both your own mother and father areound the time you were 16? IF NO: With whom were you living around that time?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab both own mother and father \tab father and stepmother \tab father only \tab mother and stepfather \tab mother only \tab other arrangement with relatives (e.g., aunt and uncle, grandparents) \tab some other female relative (no male head) \tab some other male relative (no female head) \tab no answer \tab other \tab iap \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1198 \tab 27 \tab 35 \tab 52 \tab 173 \tab 93 \tab 25 \tab 10 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1174 \tab 31 \tab 46 \tab 43 \tab 123 \tab 41 \tab 14 \tab 6 \tab 2 \tab 24 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1126 \tab 25 \tab 36 \tab 47 \tab 152 \tab 30 \tab 20 \tab 9 \tab - \tab 39 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1140 \tab 34 \tab 24 \tab 82 \tab 124 \tab 36 \tab 21 \tab 4 \tab - \tab 25 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1139 \tab 29 \tab 40 \tab 45 \tab 162 \tab 33 \tab 17 \tab 3 \tab 3 \tab 28 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1133 \tab 28 \tab 44 \tab 76 \tab 172 \tab 35 \tab 14 \tab 4 \tab 2 \tab 22 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1157 \tab 21 \tab 34 \tab 57 \tab 172 \tab 36 \tab 21 \tab 5 \tab 1 \tab 28 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1070 \tab 25 \tab 29 \tab 69 \tab 179 \tab 22 \tab 23 \tab 8 \tab - \tab 43 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1334 \tab 34 \tab 48 \tab 70 \tab 246 \tab 47 \tab 35 \tab 5 \tab 4 \tab 37 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1225 \tab 29 \tab 35 \tab 59 \tab 170 \tab 34 \tab 24 \tab 1 \tab - \tab 22 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1100 \tab 22 \tab 33 \tab 73 \tab 151 \tab 31 \tab 24 \tab 6 \tab 1 \tab 32 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1160 \tab 33 \tab 37 \tab 50 \tab 153 \tab 42 \tab 19 \tab 7 \tab - \tab 33 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1082 \tab 19 \tab 25 \tab 72 \tab 180 \tab 35 \tab 19 \tab 4 \tab - \tab 34 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1316 \tab 36 \tab 43 \tab 65 \tab 233 \tab 46 \tab 45 \tab 7 \tab 3 \tab 25 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1072 \tab 27 \tab 36 \tab 92 \tab 171 \tab 36 \tab 18 \tab 1 \tab - \tab 28 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1153 \tab 17 \tab 38 \tab 60 \tab 172 \tab 23 \tab 25 \tab 3 \tab 1 \tab 45 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 992 \tab 20 \tab 25 \tab 74 \tab 170 \tab 29 \tab 21 \tab 7 \tab 1 \tab 33 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1102 \tab 36 \tab 37 \tab 68 \tab 176 \tab 23 \tab 29 \tab 8 \tab - \tab 38 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1192 \tab 33 \tab 34 \tab 72 \tab 166 \tab 38 \tab 31 \tab 7 \tab - \tab 33 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2101 \tab 76 \tab 71 \tab 162 \tab 405 \tab 62 \tab 35 \tab 12 \tab - \tab 68 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2012 \tab 79 \tab 70 \tab 135 \tab 407 \tab 71 \tab 42 \tab 3 \tab 3 \tab 82 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1959 \tab 90 \tab 55 \tab 159 \tab 385 \tab 50 \tab 44 \tab 5 \tab - \tab 85 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1936 \tab 61 \tab 63 \tab 151 \tab 417 \tab 58 \tab 43 \tab 9 \tab - \tab 79 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1861 \tab 49 \tab 76 \tab 179 \tab 390 \tab 58 \tab 56 \tab 13 \tab 3 \tab 80 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1967 \tab 48 \tab 68 \tab 148 \tab 388 \tab 61 \tab 33 \tab 5 \tab 1 \tab 93 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2085 \tab 54 \tab 75 \tab 189 \tab 404 \tab 49 \tab 41 \tab 6 \tab 1 \tab 88 \tab 1518 \tab - \tab - \tab 4510 \cr
   2008 \tab 1375 \tab 25 \tab 50 \tab 119 \tab 300 \tab 39 \tab 36 \tab 12 \tab - \tab 67 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1358 \tab 32 \tab 57 \tab 125 \tab 319 \tab 44 \tab 35 \tab 5 \tab - \tab 69 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1303 \tab 39 \tab 53 \tab 129 \tab 294 \tab 47 \tab 39 \tab 5 \tab - \tab 65 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1672 \tab 43 \tab 62 \tab 160 \tab 410 \tab 59 \tab 44 \tab 12 \tab - \tab 76 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1845 \tab 40 \tab 80 \tab 170 \tab 489 \tab 70 \tab 41 \tab 10 \tab 2 \tab 120 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1499 \tab 37 \tab 74 \tab 150 \tab 411 \tab 56 \tab 46 \tab 9 \tab 1 \tab 65 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 2910 \tab 83 \tab 97 \tab 230 \tab 522 \tab 60 \tab 40 \tab 9 \tab 3 \tab 76 \tab - \tab 1 \tab 1 \tab 4032 \cr
   2022 \tab 2360 \tab 78 \tab 67 \tab 228 \tab 539 \tab 100 \tab 58 \tab 11 \tab 2 \tab 100 \tab - \tab - \tab 1 \tab 3544 \cr
   Total \tab 50108 \tab 1360 \tab 1697 \tab 3660 \tab 9325 \tab 1594 \tab 1078 \tab 231 \tab 34 \tab 1782 \tab 1518 \tab 1 \tab 2 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} other
\item \code{1} both own mother and father
\item \code{2} father and stepmother
\item \code{3} mother and stepfather
\item \code{4} father only
\item \code{5} mother only
\item \code{6} some other male relative (no female head)
\item \code{7} some other female relative (no male head)
\item \code{8} other arrangement with relatives (e.g., aunt and uncle, grandparents)
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
