% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{fechld}
\alias{fechld}
\title{Mother working doesn't hurt children}
\source{
General Social Survey https://gss.norc.org
}
\description{
fechld
}
\details{
Question 252. Now I'm going to read several more statements. As I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. a. A working mother can establish just as warm and secure a relationship with her children as a mother who does not work.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab strongly agree \tab strongly disagree \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 501 \tab 505 \tab 22 \tab 3 \tab 236 \tab 263 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 600 \tab 438 \tab 11 \tab 5 \tab 323 \tab 157 \tab - \tab 1534 \cr
   1986 \tab - \tab 589 \tab 434 \tab 9 \tab 1 \tab 323 \tab 114 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 384 \tab 271 \tab 9 \tab 2 \tab 233 \tab 89 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 420 \tab 286 \tab 12 \tab 4 \tab 214 \tab 70 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 382 \tab 267 \tab 6 \tab 1 \tab 201 \tab 71 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 458 \tab 287 \tab 11 \tab 2 \tab 202 \tab 64 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 496 \tab 283 \tab 15 \tab 2 \tab 217 \tab 67 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 901 \tab 496 \tab 24 \tab 5 \tab 455 \tab 96 \tab - \tab 2992 \cr
   1996 \tab 475 \tab 1012 \tab 624 \tab 29 \tab 7 \tab 579 \tab 178 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 841 \tab 470 \tab 26 \tab 5 \tab 407 \tab 122 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 752 \tab 540 \tab 22 \tab 9 \tab 376 \tab 178 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 355 \tab 244 \tab 7 \tab - \tab 218 \tab 84 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 379 \tab 244 \tab 9 \tab - \tab 211 \tab 63 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 848 \tab 533 \tab 20 \tab - \tab 474 \tab 117 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 600 \tab 293 \tab 7 \tab 2 \tab 354 \tab 73 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 648 \tab 285 \tab 5 \tab 1 \tab 425 \tab 66 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 593 \tab 291 \tab 10 \tab - \tab 342 \tab 66 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 767 \tab 323 \tab 9 \tab - \tab 499 \tab 77 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 802 \tab 395 \tab 10 \tab - \tab 613 \tab 68 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 670 \tab 324 \tab 8 \tab 1 \tab 480 \tab 76 \tab - \tab 2348 \cr
   2021 \tab 1315 \tab 1173 \tab 394 \tab - \tab - \tab 1062 \tab 85 \tab 3 \tab 4032 \cr
   2022 \tab 1168 \tab 1031 \tab 439 \tab 11 \tab 1 \tab 796 \tab 98 \tab - \tab 3544 \cr
   Total \tab 36594 \tab 15202 \tab 8666 \tab 292 \tab 51 \tab 9240 \tab 2342 \tab 3 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
