% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{fefam}
\alias{fefam}
\title{Better for man to work, woman tend home}
\source{
General Social Survey https://gss.norc.org
}
\description{
fefam
}
\details{
Question 252. Now I'm going to read several more statements. As I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. d. It is much better for everyone involved if the man is the achiever outside the home and the woman takes care of the home and family.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab strongly agree \tab strongly disagree \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 714 \tab 422 \tab 23 \tab 4 \tab 275 \tab 92 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 577 \tab 574 \tab 26 \tab 6 \tab 150 \tab 201 \tab - \tab 1534 \cr
   1986 \tab - \tab 557 \tab 576 \tab 24 \tab 2 \tab 132 \tab 179 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 318 \tab 399 \tab 19 \tab 5 \tab 88 \tab 159 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 305 \tab 413 \tab 26 \tab 3 \tab 95 \tab 164 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 301 \tab 405 \tab 26 \tab 1 \tab 64 \tab 131 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 344 \tab 405 \tab 24 \tab 4 \tab 76 \tab 171 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 318 \tab 491 \tab 23 \tab 3 \tab 64 \tab 181 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 540 \tab 896 \tab 53 \tab 8 \tab 131 \tab 349 \tab - \tab 2992 \cr
   1996 \tab 475 \tab 725 \tab 1034 \tab 64 \tab 12 \tab 179 \tab 415 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 509 \tab 836 \tab 43 \tab 10 \tab 130 \tab 343 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 531 \tab 731 \tab 53 \tab 20 \tab 204 \tab 338 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 257 \tab 382 \tab 9 \tab 1 \tab 91 \tab 168 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 257 \tab 394 \tab 11 \tab 1 \tab 79 \tab 164 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 538 \tab 918 \tab 23 \tab 1 \tab 166 \tab 346 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 359 \tab 587 \tab 19 \tab 2 \tab 122 \tab 240 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 404 \tab 610 \tab 8 \tab 4 \tab 107 \tab 297 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 313 \tab 636 \tab 15 \tab 2 \tab 87 \tab 249 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 410 \tab 785 \tab 17 \tab 3 \tab 105 \tab 355 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 405 \tab 853 \tab 15 \tab 1 \tab 110 \tab 504 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 310 \tab 714 \tab 13 \tab 1 \tab 88 \tab 433 \tab - \tab 2348 \cr
   2021 \tab 1315 \tab 410 \tab 1137 \tab - \tab - \tab 131 \tab 1030 \tab 9 \tab 4032 \cr
   2022 \tab 1168 \tab 437 \tab 1000 \tab 25 \tab 1 \tab 136 \tab 775 \tab 2 \tab 3544 \cr
   Total \tab 36594 \tab 9839 \tab 15198 \tab 559 \tab 95 \tab 2810 \tab 7284 \tab 11 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
