% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{fehelp}
\alias{fehelp}
\title{fehelp}
\source{
General Social Survey https://gss.norc.org
}
\description{
Wife should help husbands career first
}
\details{
Question 252. Now I'm going to read several more statements. As I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. b. It is more important for a wife to help her husband's career than to have one herself.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab strongly agree \tab strongly disagree \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 636 \tab 534 \tab 52 \tab 6 \tab 205 \tab 97 \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 455 \tab 679 \tab 55 \tab 7 \tab 97 \tab 241 \tab 1534 \cr
   1986 \tab - \tab 428 \tab 681 \tab 33 \tab 1 \tab 88 \tab 239 \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 253 \tab 466 \tab 21 \tab 1 \tab 47 \tab 200 \tab 1481 \cr
   1989 \tab 531 \tab 217 \tab 501 \tab 37 \tab 4 \tab 53 \tab 194 \tab 1537 \cr
   1990 \tab 444 \tab 230 \tab 457 \tab 22 \tab 2 \tab 35 \tab 182 \tab 1372 \cr
   1991 \tab 493 \tab 262 \tab 504 \tab 23 \tab 1 \tab 34 \tab 200 \tab 1517 \cr
   1993 \tab 526 \tab 214 \tab 576 \tab 33 \tab 2 \tab 33 \tab 222 \tab 1606 \cr
   1994 \tab 1015 \tab 351 \tab 1057 \tab 64 \tab 13 \tab 59 \tab 433 \tab 2992 \cr
   1996 \tab 475 \tab 420 \tab 1278 \tab 73 \tab 10 \tab 74 \tab 574 \tab 2904 \cr
   1998 \tab 961 \tab 303 \tab 999 \tab 54 \tab 12 \tab 44 \tab 459 \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 56553 \tab 3769 \tab 7732 \tab 467 \tab 59 \tab 769 \tab 3041 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
