% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_181.R
\name{fehire}
\alias{fehire}
\title{Should hire and promote women}
\source{
General Social Survey https://gss.norc.org
}
\description{
fehire
}
\details{
Question 568. Now I'm going to read several statements. As I read each one, please tell me whether you strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree: Because of past discrimination, employers should make special efforts to hire and promote qualified women.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab neither agree nor disagree \tab no answer \tab strongly agree \tab strongly disagree \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1444 \tab 553 \tab 340 \tab 22 \tab 200 \tab 2 \tab 248 \tab 95 \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1889 \tab 367 \tab 160 \tab 14 \tab 84 \tab 114 \tab 155 \tab 34 \tab - \tab 2817 \cr
   2002 \tab 2295 \tab 220 \tab 87 \tab 1 \tab 43 \tab - \tab 99 \tab 20 \tab - \tab 2765 \cr
   2004 \tab 2359 \tab 222 \tab 90 \tab 1 \tab 56 \tab 1 \tab 65 \tab 18 \tab - \tab 2812 \cr
   2006 \tab 3512 \tab 463 \tab 183 \tab 4 \tab 133 \tab - \tab 181 \tab 34 \tab - \tab 4510 \cr
   2008 \tab 1342 \tab 316 \tab 153 \tab 3 \tab 66 \tab 1 \tab 120 \tab 22 \tab - \tab 2023 \cr
   2010 \tab 1345 \tab 342 \tab 126 \tab 2 \tab 71 \tab 1 \tab 127 \tab 30 \tab - \tab 2044 \cr
   2012 \tab 1332 \tab 314 \tab 132 \tab 4 \tab 68 \tab 1 \tab 102 \tab 21 \tab - \tab 1974 \cr
   2014 \tab 1689 \tab 432 \tab 157 \tab 4 \tab 75 \tab 2 \tab 156 \tab 23 \tab - \tab 2538 \cr
   2016 \tab 1915 \tab 483 \tab 141 \tab 1 \tab 87 \tab - \tab 210 \tab 30 \tab - \tab 2867 \cr
   2018 \tab 1583 \tab 380 \tab 105 \tab 4 \tab 75 \tab 1 \tab 176 \tab 24 \tab - \tab 2348 \cr
   2021 \tab 1315 \tab 898 \tab 381 \tab 1 \tab 673 \tab - \tab 622 \tab 134 \tab 8 \tab 4032 \cr
   2022 \tab 1168 \tab 955 \tab 334 \tab 15 \tab 417 \tab 2 \tab 556 \tab 95 \tab 2 \tab 3544 \cr
   Total \tab 58400 \tab 5945 \tab 2389 \tab 76 \tab 2048 \tab 125 \tab 2817 \tab 580 \tab 10 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
