% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_177.R
\name{fejobaff}
\alias{fejobaff}
\title{fejobaff}
\source{
General Social Survey https://gss.norc.org
}
\description{
For or against preferential hiring of women
}
\details{
Question 552. Some people say that because of past discrimination, women should be given preference in hiring and promotion. Others say that such preference in hiring and promotion of women is wrong because it discriminates against men. What about you opinion - are you for or against preferential hiring and promotion of women? Do you favor/opp;ose preference in hiring and promotion strongly or not strongly?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab don't know \tab no answer \tab not strongly favor \tab not strongly oppose \tab strongly favor \tab strongly oppose \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1913 \tab 49 \tab 9 \tab 95 \tab 240 \tab 152 \tab 446 \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1869 \tab 62 \tab 45 \tab 101 \tab 251 \tab 175 \tab 314 \tab - \tab 2817 \cr
   2002 \tab 2327 \tab 13 \tab - \tab 42 \tab 124 \tab 87 \tab 172 \tab - \tab 2765 \cr
   2004 \tab 2359 \tab 10 \tab 2 \tab 40 \tab 123 \tab 69 \tab 209 \tab - \tab 2812 \cr
   2006 \tab 3516 \tab 32 \tab - \tab 98 \tab 265 \tab 230 \tab 369 \tab - \tab 4510 \cr
   2008 \tab 1375 \tab 32 \tab 6 \tab 64 \tab 168 \tab 140 \tab 238 \tab - \tab 2023 \cr
   2010 \tab 1313 \tab 35 \tab 3 \tab 85 \tab 186 \tab 140 \tab 282 \tab - \tab 2044 \cr
   2012 \tab 1314 \tab 30 \tab 2 \tab 74 \tab 173 \tab 143 \tab 238 \tab - \tab 1974 \cr
   2014 \tab 1712 \tab 32 \tab 4 \tab 85 \tab 253 \tab 180 \tab 272 \tab - \tab 2538 \cr
   2016 \tab 1931 \tab 37 \tab 3 \tab 99 \tab 265 \tab 216 \tab 316 \tab - \tab 2867 \cr
   2018 \tab 1554 \tab 39 \tab 5 \tab 79 \tab 215 \tab 207 \tab 249 \tab - \tab 2348 \cr
   2021 \tab 2699 \tab 10 \tab - \tab 235 \tab 352 \tab 257 \tab 479 \tab - \tab 4032 \cr
   2022 \tab 2339 \tab 37 \tab 7 \tab 214 \tab 291 \tab 303 \tab 352 \tab 1 \tab 3544 \cr
   Total \tab 61433 \tab 418 \tab 86 \tab 1311 \tab 2906 \tab 2299 \tab 3936 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} not strongly favor
\item \code{3} not strongly oppose
\item \code{4} strongly oppose
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
