% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_66.R
\name{fepres}
\alias{fepres}
\title{fepres}
\source{
General Social Survey https://gss.norc.org
}
\description{
Vote for woman president
}
\details{
Question 201. If your party nominated a woman for President, would you vote for her if she were qualified for the job?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab don't know \tab no \tab no answer \tab yes \tab iap \tab wouldnt vote \tab Total \cr
   1972 \tab 78 \tab 404 \tab 2 \tab 1129 \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab - \tab - \tab - \tab 1504 \tab - \tab 1504 \cr
   1974 \tab 46 \tab 283 \tab 5 \tab 1150 \tab - \tab - \tab 1484 \cr
   1975 \tab 46 \tab 282 \tab 1 \tab 1158 \tab - \tab 3 \tab 1490 \cr
   1976 \tab - \tab - \tab - \tab - \tab 1499 \tab - \tab 1499 \cr
   1977 \tab 42 \tab 307 \tab 4 \tab 1177 \tab - \tab - \tab 1530 \cr
   1978 \tab 40 \tab 275 \tab - \tab 1217 \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab - \tab - \tab 1468 \tab - \tab 1468 \cr
   1982 \tab 61 \tab 249 \tab 3 \tab 1547 \tab - \tab - \tab 1860 \cr
   1983 \tab 49 \tab 208 \tab 4 \tab 1337 \tab - \tab 1 \tab 1599 \cr
   1984 \tab - \tab - \tab - \tab - \tab 1473 \tab - \tab 1473 \cr
   1985 \tab 47 \tab 262 \tab 6 \tab 1219 \tab - \tab - \tab 1534 \cr
   1986 \tab 39 \tab 195 \tab 4 \tab 1232 \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab - \tab - \tab 1819 \tab - \tab 1819 \cr
   1988 \tab 27 \tab 116 \tab 3 \tab 842 \tab 493 \tab - \tab 1481 \cr
   1989 \tab 50 \tab 129 \tab 3 \tab 824 \tab 531 \tab - \tab 1537 \cr
   1990 \tab 30 \tab 94 \tab 3 \tab 801 \tab 444 \tab - \tab 1372 \cr
   1991 \tab 43 \tab 92 \tab 4 \tab 885 \tab 493 \tab - \tab 1517 \cr
   1993 \tab 45 \tab 96 \tab 3 \tab 936 \tab 526 \tab - \tab 1606 \cr
   1994 \tab 51 \tab 154 \tab 6 \tab 1766 \tab 1015 \tab - \tab 2992 \cr
   1996 \tab 52 \tab 132 \tab 5 \tab 1771 \tab 944 \tab - \tab 2904 \cr
   1998 \tab 63 \tab 116 \tab 5 \tab 1687 \tab 961 \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab 4510 \cr
   2008 \tab 17 \tab 81 \tab 7 \tab 1224 \tab 694 \tab - \tab 2023 \cr
   2010 \tab 13 \tab 56 \tab 6 \tab 1355 \tab 614 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab 3544 \cr
   Total \tab 839 \tab 3531 \tab 74 \tab 23257 \tab 44685 \tab 4 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{5} wouldnt vote
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
