% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_79.R
\name{fepresch}
\alias{fepresch}
\title{fepresch}
\source{
General Social Survey https://gss.norc.org
}
\description{
Preschool kids suffer if mother works
}
\details{
Question 252. Now I'm going to read several more statements. As I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. c. A preschool child is likely to suffer if his or her mother works.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab strongly agree \tab strongly disagree \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 697 \tab 422 \tab 28 \tab 4 \tab 311 \tab 68 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 614 \tab 544 \tab 28 \tab 7 \tab 195 \tab 146 \tab - \tab 1534 \cr
   1986 \tab - \tab 582 \tab 572 \tab 27 \tab 1 \tab 156 \tab 132 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 361 \tab 394 \tab 19 \tab 4 \tab 104 \tab 106 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 380 \tab 411 \tab 27 \tab 5 \tab 88 \tab 95 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 372 \tab 377 \tab 26 \tab 2 \tab 73 \tab 78 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 385 \tab 421 \tab 24 \tab 4 \tab 91 \tab 99 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 377 \tab 493 \tab 28 \tab 4 \tab 75 \tab 103 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 676 \tab 898 \tab 32 \tab 9 \tab 146 \tab 216 \tab - \tab 2992 \cr
   1996 \tab 475 \tab 881 \tab 1003 \tab 74 \tab 11 \tab 203 \tab 257 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 611 \tab 863 \tab 53 \tab 9 \tab 155 \tab 180 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 682 \tab 761 \tab 56 \tab 15 \tab 175 \tab 188 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 322 \tab 392 \tab 10 \tab - \tab 88 \tab 96 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 316 \tab 408 \tab 15 \tab 1 \tab 58 \tab 108 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 622 \tab 959 \tab 24 \tab - \tab 172 \tab 215 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 370 \tab 665 \tab 19 \tab 2 \tab 104 \tab 169 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 407 \tab 718 \tab 14 \tab 3 \tab 87 \tab 201 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 350 \tab 652 \tab 18 \tab - \tab 86 \tab 196 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 420 \tab 867 \tab 12 \tab 2 \tab 82 \tab 292 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 448 \tab 1000 \tab 13 \tab - \tab 72 \tab 355 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 347 \tab 820 \tab 21 \tab 2 \tab 81 \tab 288 \tab - \tab 2348 \cr
   2021 \tab 1315 \tab 520 \tab 1433 \tab 5 \tab - \tab 105 \tab 649 \tab 5 \tab 4032 \cr
   2022 \tab 1168 \tab 514 \tab 1230 \tab 25 \tab 1 \tab 110 \tab 494 \tab 2 \tab 3544 \cr
   Total \tab 36594 \tab 11254 \tab 16303 \tab 598 \tab 86 \tab 2817 \tab 4731 \tab 7 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
