% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_552.R
\name{frndsex}
\alias{frndsex}
\title{R had sex with friend last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
frndsex
}
\details{
Question 1543. If you had other partners, please indicate all categories that apply to them. a. Close personal friend
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab don't know \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1243 \tab 59 \tab 68 \tab 111 \tab - \tab 1481 \cr
   1989 \tab 1303 \tab 49 \tab 74 \tab 111 \tab - \tab 1537 \cr
   1990 \tab 1155 \tab 58 \tab 71 \tab 88 \tab - \tab 1372 \cr
   1991 \tab 1293 \tab 65 \tab 67 \tab 92 \tab - \tab 1517 \cr
   1993 \tab 1359 \tab 73 \tab 87 \tab 87 \tab - \tab 1606 \cr
   1994 \tab 2489 \tab 143 \tab 169 \tab 191 \tab - \tab 2992 \cr
   1996 \tab 2401 \tab 161 \tab 119 \tab 223 \tab - \tab 2904 \cr
   1998 \tab 2383 \tab 139 \tab 154 \tab 156 \tab - \tab 2832 \cr
   2000 \tab 2335 \tab 128 \tab 177 \tab 177 \tab - \tab 2817 \cr
   2002 \tab 2317 \tab 104 \tab 49 \tab 294 \tab 1 \tab 2765 \cr
   2004 \tab 2429 \tab 70 \tab 36 \tab 277 \tab - \tab 2812 \cr
   2006 \tab 4087 \tab 97 \tab 16 \tab 310 \tab - \tab 4510 \cr
   2008 \tab 1688 \tab 113 \tab 23 \tab 199 \tab - \tab 2023 \cr
   2010 \tab 1706 \tab 99 \tab 30 \tab 208 \tab 1 \tab 2044 \cr
   2012 \tab 1669 \tab 98 \tab 34 \tab 173 \tab - \tab 1974 \cr
   2014 \tab 2144 \tab 136 \tab 52 \tab 206 \tab - \tab 2538 \cr
   2016 \tab 2530 \tab 112 \tab 20 \tab 205 \tab - \tab 2867 \cr
   2018 \tab 2130 \tab 70 \tab 19 \tab 127 \tab 2 \tab 2348 \cr
   2021 \tab 3747 \tab 111 \tab 20 \tab 146 \tab 8 \tab 4032 \cr
   2022 \tab 3288 \tab 62 \tab 41 \tab 131 \tab 22 \tab 3544 \cr
   Total \tab 65571 \tab 1947 \tab 1326 \tab 3512 \tab 34 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
