% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_583.R
\name{gender3}
\alias{gender3}
\title{gender3}
\source{
General Social Survey https://gss.norc.org
}
\description{
Gender of 3rd person
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. b. Code sex \link{ASK IF NOT OBVIOUS}
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab female \tab male \tab no answer \tab don't know \tab not available in this release \tab 3 \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 656 \tab 371 \tab 463 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1134 \tab 164 \tab 200 \tab 1 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 724 \tab 368 \tab 436 \tab 2 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 749 \tab 361 \tab 419 \tab 3 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 780 \tab 322 \tab 362 \tab 4 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 991 \tab 445 \tab 416 \tab 8 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 797 \tab 400 \tab 394 \tab 8 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 783 \tab 341 \tab 345 \tab 4 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 844 \tab 336 \tab 352 \tab 2 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 768 \tab 347 \tab 354 \tab 1 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 948 \tab 421 \tab 447 \tab 3 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 847 \tab 313 \tab 320 \tab 1 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 834 \tab 305 \tab 386 \tab 12 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 810 \tab 263 \tab 293 \tab 6 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 849 \tab 317 \tab 348 \tab 3 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 895 \tab 344 \tab 364 \tab 3 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1751 \tab 569 \tab 656 \tab 16 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1729 \tab 566 \tab 574 \tab 35 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1713 \tab 546 \tab 557 \tab 16 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1668 \tab 530 \tab 582 \tab 37 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1815 \tab 462 \tab 488 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1780 \tab 472 \tab 553 \tab 7 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2778 \tab 816 \tab 901 \tab 2 \tab 13 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1221 \tab 391 \tab 401 \tab 3 \tab 7 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1319 \tab 352 \tab 364 \tab 4 \tab 5 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1259 \tab 328 \tab 376 \tab 8 \tab 3 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1642 \tab 403 \tab 481 \tab 6 \tab 6 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1919 \tab 417 \tab 513 \tab 10 \tab 8 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1561 \tab 367 \tab 405 \tab 10 \tab 5 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab 4032 \cr
   2022 \tab 2841 \tab 335 \tab 355 \tab 4 \tab - \tab - \tab 7 \tab 2 \tab 3544 \cr
   Total \tab 43006 \tab 11972 \tab 13105 \tab 219 \tab 47 \tab 4032 \tab 7 \tab 2 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
