% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_583.R
\name{gender4}
\alias{gender4}
\title{Gender of 4th person}
\source{
General Social Survey https://gss.norc.org
}
\description{
gender4
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. b. Code sex \link{ASK IF NOT OBVIOUS}
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab female \tab male \tab no answer \tab don't know \tab not available in this release \tab 3 \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 952 \tab 254 \tab 284 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1255 \tab 115 \tab 128 \tab 1 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1007 \tab 272 \tab 250 \tab 1 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1017 \tab 242 \tab 271 \tab 2 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1040 \tab 209 \tab 217 \tab 2 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1343 \tab 271 \tab 243 \tab 3 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1104 \tab 244 \tab 247 \tab 4 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1062 \tab 182 \tab 226 \tab 3 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1112 \tab 208 \tab 213 \tab 1 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1032 \tab 199 \tab 238 \tab 1 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1282 \tab 254 \tab 280 \tab 3 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1097 \tab 191 \tab 193 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1098 \tab 198 \tab 231 \tab 10 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1032 \tab 179 \tab 158 \tab 3 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1125 \tab 187 \tab 203 \tab 2 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1190 \tab 192 \tab 221 \tab 3 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2274 \tab 342 \tab 369 \tab 7 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2184 \tab 359 \tab 342 \tab 19 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2195 \tab 316 \tab 309 \tab 12 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2127 \tab 334 \tab 338 \tab 18 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2266 \tab 251 \tab 247 \tab 1 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2214 \tab 270 \tab 320 \tab 8 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 3502 \tab 459 \tab 534 \tab 3 \tab 12 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1547 \tab 225 \tab 238 \tab 4 \tab 9 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1627 \tab 182 \tab 225 \tab 4 \tab 6 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1526 \tab 222 \tab 219 \tab 5 \tab 2 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2040 \tab 230 \tab 253 \tab 7 \tab 8 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2353 \tab 231 \tab 267 \tab 10 \tab 6 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1899 \tab 217 \tab 223 \tab 6 \tab 3 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab 4032 \cr
   2022 \tab 3136 \tab 206 \tab 197 \tab - \tab - \tab - \tab 4 \tab 1 \tab 3544 \cr
   Total \tab 53239 \tab 7241 \tab 7684 \tab 143 \tab 46 \tab 4032 \tab 4 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
