% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_584.R
\name{gender6}
\alias{gender6}
\title{Gender of 6th person}
\source{
General Social Survey https://gss.norc.org
}
\description{
gender6
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. b. Code sex \link{ASK IF NOT OBVIOUS}
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab female \tab male \tab no answer \tab don't know \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1363 \tab 62 \tab 65 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1452 \tab 23 \tab 23 \tab 1 \tab - \tab - \tab 1499 \cr
   1977 \tab 1413 \tab 61 \tab 55 \tab 1 \tab - \tab - \tab 1530 \cr
   1978 \tab 1424 \tab 54 \tab 53 \tab 1 \tab - \tab - \tab 1532 \cr
   1980 \tab 1392 \tab 36 \tab 38 \tab 2 \tab - \tab - \tab 1468 \cr
   1982 \tab 1763 \tab 49 \tab 47 \tab 1 \tab - \tab - \tab 1860 \cr
   1983 \tab 1513 \tab 38 \tab 47 \tab 1 \tab - \tab - \tab 1599 \cr
   1984 \tab 1402 \tab 31 \tab 40 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1472 \tab 30 \tab 32 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1402 \tab 32 \tab 35 \tab 1 \tab - \tab - \tab 1470 \cr
   1987 \tab 1727 \tab 35 \tab 57 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1428 \tab 29 \tab 23 \tab 1 \tab - \tab - \tab 1481 \cr
   1989 \tab 1464 \tab 36 \tab 33 \tab 4 \tab - \tab - \tab 1537 \cr
   1990 \tab 1325 \tab 21 \tab 26 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1468 \tab 17 \tab 32 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1566 \tab 21 \tab 19 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2900 \tab 46 \tab 44 \tab 2 \tab - \tab - \tab 2992 \cr
   1996 \tab 2791 \tab 55 \tab 55 \tab 3 \tab - \tab - \tab 2904 \cr
   1998 \tab 2741 \tab 40 \tab 49 \tab 2 \tab - \tab - \tab 2832 \cr
   2000 \tab 2721 \tab 44 \tab 46 \tab 6 \tab - \tab - \tab 2817 \cr
   2002 \tab 2701 \tab 30 \tab 34 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2724 \tab 42 \tab 46 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4358 \tab 72 \tab 76 \tab 1 \tab 3 \tab - \tab 4510 \cr
   2008 \tab 1950 \tab 26 \tab 46 \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab 1950 \tab 44 \tab 46 \tab 2 \tab 2 \tab - \tab 2044 \cr
   2012 \tab 1898 \tab 45 \tab 30 \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab 2463 \tab 31 \tab 43 \tab - \tab 1 \tab - \tab 2538 \cr
   2016 \tab 2770 \tab 46 \tab 46 \tab 3 \tab 2 \tab - \tab 2867 \cr
   2018 \tab 2294 \tab 24 \tab 29 \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3465 \tab 47 \tab 32 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 65901 \tab 1167 \tab 1247 \tab 35 \tab 8 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
