% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gssr-package.R
\name{get_asked}
\alias{get_asked}
\title{Get whether a question was asked}
\usage{
get_asked(x)
}
\arguments{
\item{x}{The GSS variable being assessed}
}
\value{
TRUE or FALSE depending on whether it passes the test
}
\description{
Get whether a question was asked
}
\examples{
# See gss_which_years()
}
\keyword{internal}
