% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_116.R
\name{god}
\alias{god}
\title{god}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's confidence in the existence of god
}
\details{
Question 369. Please look at this card and tell me which statement comes closest to expresing what you believe about God: 1. I don't believe in God. 2. I don't know whether there is a God and I don't believe there is any way to find out. 3. I don't believe in a personal God, but I do believe in a Higher Power of some kind. 4. I find myself believing in God some of the time, but not at others. 5. While I have doubts, I feel that I do believe in God. 6. I know God really exists and I have no doubts about it.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab believe sometimes \tab believe with doubts \tab don't believe \tab don't know, no way to find out \tab higher power \tab no answer \tab no doubts \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 65 \tab 279 \tab 22 \tab 56 \tab 117 \tab 3 \tab 939 \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 158 \tab 69 \tab 241 \tab 31 \tab 60 \tab 92 \tab 31 \tab 834 \tab 1 \tab - \tab 1517 \cr
   1993 \tab 49 \tab 47 \tab 224 \tab 44 \tab 63 \tab 127 \tab 37 \tab 992 \tab 23 \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 51 \tab 214 \tab 33 \tab 37 \tab 131 \tab 34 \tab 860 \tab 26 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1548 \tab 58 \tab 181 \tab 40 \tab 60 \tab 121 \tab 48 \tab 775 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 1541 \tab 41 \tab 192 \tab 33 \tab 49 \tab 83 \tab 66 \tab 766 \tab 46 \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 128 \tab 473 \tab 66 \tab 129 \tab 294 \tab 13 \tab 1876 \tab 13 \tab - \tab 4510 \cr
   2008 \tab - \tab 76 \tab 330 \tab 60 \tab 95 \tab 204 \tab 9 \tab 1240 \tab 9 \tab - \tab 2023 \cr
   2010 \tab - \tab 93 \tab 344 \tab 66 \tab 123 \tab 216 \tab 7 \tab 1181 \tab 14 \tab - \tab 2044 \cr
   2012 \tab - \tab 81 \tab 307 \tab 65 \tab 107 \tab 228 \tab 10 \tab 1168 \tab 8 \tab - \tab 1974 \cr
   2014 \tab - \tab 91 \tab 416 \tab 87 \tab 143 \tab 316 \tab 11 \tab 1468 \tab 6 \tab - \tab 2538 \cr
   2016 \tab - \tab 100 \tab 461 \tab 126 \tab 188 \tab 357 \tab 19 \tab 1607 \tab 9 \tab - \tab 2867 \cr
   2018 \tab - \tab 88 \tab 410 \tab 109 \tab 150 \tab 298 \tab 14 \tab 1262 \tab 17 \tab - \tab 2348 \cr
   2021 \tab 1389 \tab 124 \tab 427 \tab 173 \tab 237 \tab 341 \tab 1 \tab 1313 \tab 2 \tab 25 \tab 4032 \cr
   2022 \tab 1176 \tab 125 \tab 382 \tab 158 \tab 176 \tab 336 \tab 15 \tab 1159 \tab 6 \tab 11 \tab 3544 \cr
   Total \tab 42250 \tab 1237 \tab 4881 \tab 1113 \tab 1673 \tab 3261 \tab 318 \tab 17440 \tab 181 \tab 36 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} don't believe
\item \code{2} don't know, no way to find out
\item \code{3} higher power
\item \code{4} believe sometimes
\item \code{5} believe with doubts
\item \code{6} no doubts
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
