% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_468.R
\name{goodlife}
\alias{goodlife}
\title{Standard of living of r will improve}
\source{
General Social Survey https://gss.norc.org
}
\description{
goodlife
}
\details{
Question 1251. The way things are in America, people like me and my family have a good chance of improving our standard of living. Do you agree or disagree?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab agree \tab can't choose \tab disagree \tab neither agree nor disagree \tab no answer \tab strongly agree \tab strongly disagree \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 255 \tab 791 \tab 29 \tab 135 \tab 271 \tab 26 \tab 282 \tab 30 \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1492 \tab 757 \tab 27 \tab 317 \tab 176 \tab 6 \tab 163 \tab 54 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 973 \tab 25 \tab 388 \tab 211 \tab 5 \tab 251 \tab 72 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 1040 \tab 30 \tab 229 \tab 211 \tab 1 \tab 350 \tab 50 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 1003 \tab 14 \tab 217 \tab 171 \tab 5 \tab 439 \tab 47 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 447 \tab 3 \tab 118 \tab 96 \tab - \tab 231 \tab 17 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 423 \tab 2 \tab 144 \tab 118 \tab 1 \tab 158 \tab 30 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 931 \tab 10 \tab 275 \tab 303 \tab - \tab 413 \tab 57 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 572 \tab 10 \tab 323 \tab 203 \tab - \tab 206 \tab 51 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 611 \tab 7 \tab 300 \tab 227 \tab 2 \tab 178 \tab 52 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 549 \tab 6 \tab 306 \tab 224 \tab - \tab 182 \tab 71 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 743 \tab 3 \tab 353 \tab 296 \tab 1 \tab 230 \tab 61 \tab - \tab 2538 \cr
   2016 \tab 912 \tab 832 \tab 3 \tab 390 \tab 334 \tab 2 \tab 292 \tab 102 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 716 \tab 8 \tab 230 \tab 291 \tab 2 \tab 271 \tab 45 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 944 \tab 2 \tab 526 \tab 750 \tab 1 \tab 290 \tab 154 \tab 5 \tab 4032 \cr
   2022 \tab 1173 \tab 823 \tab 11 \tab 497 \tab 557 \tab 2 \tab 285 \tab 191 \tab 5 \tab 3544 \cr
   Total \tab 45489 \tab 12155 \tab 190 \tab 4748 \tab 4439 \tab 54 \tab 4221 \tab 1084 \tab 10 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
