% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_14.R
\name{granborn}
\alias{granborn}
\title{granborn}
\source{
General Social Survey https://gss.norc.org
}
\description{
How many grandparents born outside u.s.
}
\details{
Question 33. Were all of your four grandparents born in this country? a. How many were born outside the United States?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab don't know \tab four \tab no answer \tab none \tab one \tab three \tab two \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 35 \tab 20 \tab 337 \tab 5 \tab 840 \tab 81 \tab 30 \tab 182 \tab - \tab 1530 \cr
   1978 \tab 50 \tab 38 \tab 299 \tab 4 \tab 845 \tab 101 \tab 40 \tab 155 \tab - \tab 1532 \cr
   1980 \tab 44 \tab 28 \tab 302 \tab 8 \tab 793 \tab 91 \tab 39 \tab 163 \tab - \tab 1468 \cr
   1982 \tab 83 \tab 12 \tab 344 \tab 8 \tab 1079 \tab 89 \tab 50 \tab 195 \tab - \tab 1860 \cr
   1983 \tab 82 \tab 16 \tab 361 \tab 6 \tab 786 \tab 104 \tab 52 \tab 192 \tab - \tab 1599 \cr
   1984 \tab 77 \tab 7 \tab 283 \tab 12 \tab 802 \tab 94 \tab 30 \tab 168 \tab - \tab 1473 \cr
   1985 \tab 53 \tab 18 \tab 291 \tab 8 \tab 832 \tab 117 \tab 35 \tab 180 \tab - \tab 1534 \cr
   1986 \tab 73 \tab 9 \tab 306 \tab 12 \tab 777 \tab 98 \tab 42 \tab 153 \tab - \tab 1470 \cr
   1987 \tab 91 \tab 24 \tab 261 \tab 11 \tab 1143 \tab 90 \tab 27 \tab 172 \tab - \tab 1819 \cr
   1988 \tab 103 \tab 13 \tab 262 \tab 3 \tab 810 \tab 92 \tab 41 \tab 157 \tab - \tab 1481 \cr
   1989 \tab 76 \tab 21 \tab 279 \tab 11 \tab 859 \tab 92 \tab 37 \tab 162 \tab - \tab 1537 \cr
   1990 \tab 68 \tab 13 \tab 244 \tab 8 \tab 763 \tab 87 \tab 37 \tab 152 \tab - \tab 1372 \cr
   1991 \tab 74 \tab 17 \tab 272 \tab 14 \tab 843 \tab 93 \tab 44 \tab 160 \tab - \tab 1517 \cr
   1993 \tab 83 \tab 14 \tab 272 \tab 11 \tab 925 \tab 96 \tab 39 \tab 166 \tab - \tab 1606 \cr
   1994 \tab 159 \tab 19 \tab 503 \tab 15 \tab 1747 \tab 187 \tab 57 \tab 305 \tab - \tab 2992 \cr
   1996 \tab 157 \tab 21 \tab 466 \tab 16 \tab 1672 \tab 176 \tab 68 \tab 328 \tab - \tab 2904 \cr
   1998 \tab 153 \tab 21 \tab 486 \tab 28 \tab 1606 \tab 165 \tab 63 \tab 310 \tab - \tab 2832 \cr
   2000 \tab 174 \tab 25 \tab 476 \tab 17 \tab 1605 \tab 152 \tab 61 \tab 307 \tab - \tab 2817 \cr
   2002 \tab 134 \tab 14 \tab 492 \tab 8 \tab 1626 \tab 153 \tab 57 \tab 281 \tab - \tab 2765 \cr
   2004 \tab 116 \tab 19 \tab 525 \tab 2 \tab 1652 \tab 169 \tab 55 \tab 274 \tab - \tab 2812 \cr
   2006 \tab 1639 \tab 18 \tab 649 \tab - \tab 1676 \tab 181 \tab 72 \tab 275 \tab - \tab 4510 \cr
   2008 \tab 96 \tab 7 \tab 409 \tab - \tab 1149 \tab 128 \tab 40 \tab 194 \tab - \tab 2023 \cr
   2010 \tab 111 \tab 19 \tab 412 \tab 3 \tab 1152 \tab 115 \tab 38 \tab 194 \tab - \tab 2044 \cr
   2012 \tab 104 \tab 11 \tab 409 \tab 6 \tab 1087 \tab 121 \tab 34 \tab 202 \tab - \tab 1974 \cr
   2014 \tab 126 \tab 13 \tab 557 \tab 1 \tab 1356 \tab 160 \tab 53 \tab 272 \tab - \tab 2538 \cr
   2016 \tab 136 \tab 20 \tab 564 \tab 8 \tab 1609 \tab 197 \tab 49 \tab 284 \tab - \tab 2867 \cr
   2018 \tab 121 \tab 11 \tab 461 \tab 2 \tab 1333 \tab 124 \tab 60 \tab 236 \tab - \tab 2348 \cr
   2021 \tab 216 \tab 108 \tab 585 \tab 62 \tab 2379 \tab 220 \tab 99 \tab 350 \tab 13 \tab 4032 \cr
   2022 \tab 216 \tab 62 \tab 646 \tab 15 \tab 2012 \tab 206 \tab 89 \tab 296 \tab 2 \tab 3544 \cr
   Total \tab 12240 \tab 638 \tab 11753 \tab 304 \tab 35758 \tab 3779 \tab 1438 \tab 6465 \tab 15 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} none
\item \code{1} one
\item \code{2} two
\item \code{3} three
\item \code{4} four
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
