% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_all}
\alias{gss_all}
\title{General Social Survey Survey Cumulative Data File 1972-2022}
\format{
An object of class `tibble` with 72,390 rows and 6,693 columns. Variables are encoded as labelled vectors. The GSS Codebook is the authoritative source for the variables in this dataset. It is available at <http://gss.norc.org/Get-Documentation>. Summary information is available in `gss_doc`, a tibble supplied with this package.
}
\source{
National Opinion Research Center, [http://gss.norc.org](http://gss.norc.org).
}
\usage{
data(gss_all)
}
\description{
A tibble containing the General Social Survey Cumulative Data file.
}
\keyword{datasets}
