% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_dict}
\alias{gss_dict}
\title{Data Dictionary for the GSS Cumulative Data File 1972-2018}
\format{
A tibble with  rows and  columns.
\describe{
\item{\code{pos}}{Column-index position of the variable in \code{gss_all}. (Integer.)}
\item{\code{variable}}{The name of the variable. (Character.)}
\item{\code{label}}{The variable's label. 'r' means 'Respondent'. (Character.)}
\item{\code{col_type}}{The column type of the variable in \code{gss_all}. (Character.)}
\item{\code{value_labels}}{The value labels and associated numeric values for the variable. Numeric values are given in square brackets. (Character.)}
\item{\code{years}}{Which years the question was asked. This information is stored as a tibble in a list column. The tibble contains a column for the \code{year} and a TRUE/FALSE coumn for \code{present}.}
}
}
\source{
National Opinion Research Center, \href{http://gss.norc.org}{http://gss.norc.org}.
}
\usage{
data(gss_dict)
}
\description{
A tibble containing information on the variables in the GSS
Cumulative Data File. This is a more compact version of the \code{gss_doc} codebook.
See http://gss.norc.org/Get-Documentation for full documentation of the variables.
}
\keyword{datasets}
