% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_doc}
\alias{gss_doc}
\title{Codebook for the GSS Cumulative Data File 1972-2022}
\format{
A tibble.
\describe{
\item{\code{id}}{The short name of the variable. A character vector.}
\item{\code{description}}{Brief description of the variable. A
character vector.}
\item{\code{properties}}{A list column containing information on
the variable's data type, its missing value codes, and its
record/column location.}
\item{\code{marginals}}{A list column containing a summary of the
values taken by the variable.}
\item{\code{text}}{Further more detailed information on the
question asked and the directions provided.}
}
}
\source{
National Opinion Research Center, \url{http://gss.norc.org}.
}
\usage{
data(gss_doc)
}
\description{
A tibble containing information on the variables in the GSS Cumulative Data File.
}
\keyword{datasets}
