% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gssr-package.R
\name{gss_get_props}
\alias{gss_get_props}
\title{Property information for a GSS variable or variables}
\usage{
gss_get_props(varnames = "id", data = gss_doc, props = properties)
}
\arguments{
\item{varnames}{The variable or variables you want
to see properties for. Can be a character string or a character
vector.}

\item{data}{The codebook data frame, by default \code{gss_doc}}

\item{props}{Which properties. Not needed for \code{gss_doc}, but one of \code{properties_1}, \code{properties_2}, or \code{properties_3} when \code{data} is \code{gss_panel_docw}.}
}
\value{
A tibble of the properties for each variable
}
\description{
Property information for a GSS variable or variables
}
\details{
Returns the properties of a GSS variable as given in the codebook,
typically the Data Type, Missing Data Codes, and Record/Column
location.
}
\examples{
\donttest{
data(gss_doc)
gss_get_props(varnames = "age")

gss_get_props(varnames = c("age", "race", "fefam"))
}
}
\author{
Kieran Healy
}
