% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gssr-package.R
\name{gss_get_years}
\alias{gss_get_years}
\title{(DEPRECATED) What years was a particular question asked in the GSS?}
\usage{
gss_get_years()
}
\description{
This function is deprecated. Use `gss_which_years()` instead.
}
\examples{
\donttest{
}
}
\keyword{internal}
