% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gssr-package.R
\name{gss_get_yr}
\alias{gss_get_yr}
\title{Download GSS data file for a single year from NORC}
\usage{
gss_get_yr(
  year = 2018,
  url = "http://www.gss.norc.org/documents/stata/",
  fname = "_stata",
  ext = "zip",
  dest = "data-raw/",
  save_file = c("n", "y")
)
}
\arguments{
\item{year}{The desired GSS survey year}

\item{url}{Location of the file. Defaults to the current NORC URL}

\item{fname}{Non-year filename component. Defaults to '_stata'. Usually should not be changed.}

\item{ext}{File name extension. Defaults to 'zip'. Usually should not be changed.}

\item{dest}{If \code{save_file} is "y", the directory to put the file in. Defaults to \code{data-raw} in current directory.}

\item{save_file}{Save the data file as well as loading it as an object. Defaults to 'n'.}
}
\value{
A tibble with the requested year's GSS data.
}
\description{
Use \code{gss_get_yr()} to get GSS data for a single year from
NORC's GSS website (where it is available as a zipped Stata file)
and put it directly into a tibble.
}
\examples{
\donttest{
gss80 <- gss_get_yr(1980)
}
}
