% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_panel06_long}
\alias{gss_panel06_long}
\title{General Social Survey Survey 2006 Three Wave Panel Data}
\format{
A tibble with 6,000 rows and 1,572 columns. Variables are encoded as labelled vectors. The GSS
Codebook is the authoritative source for the variables in this
dataset. It is available at
\url{http://gss.norc.org/Get-Documentation}. Summary
information is available in \code{gss_panel_doc}, a tibble
supplied with this package. Respondent ids are contained in the
variable \code{firstid} (from the GSS \verb{id\\_1} variable).
Survey waves (years 2006, 2008, 2010) are contained in the
\code{wave} variable as 1, 2, and 3. See also the \code{gss_panel_doc} object in this package.
}
\source{
National Opinion Research Center, \url{http://gss.norc.org}.
}
\usage{
data(gss_panel06_long)
}
\description{
A tibble containing the General Social Survey 2006 Three Wave Panel Data
File, in long format.
}
\keyword{datasets}
