% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_panel_doc}
\alias{gss_panel_doc}
\title{Codebook for the GSS Three Wave Panel}
\format{
A tibble with 628 rows and nine columns.
\describe{
\item{`id`}{The short name of the variable, without its wave identifier. A character vector.}
\item{`description`}{Brief description of the variable. A
character vector.}
\item{`text`}{Further more detailed information on the
question asked and the directions provided.}
\item{`properties_1`}{A list column containing information on
the variable's data type, its missing value codes, and its
record/column location for Wave 1 of the panel.}
\item{`properties_2`}{A list column containing information on
the variable's data type, its missing value codes, and its
record/column location for Wave 2 of the panel.}
\item{`properties_3`}{A list column containing information on
the variable's data type, its missing value codes, and its
record/column location for Wave 3 of the panel.}
\item{`marginals_1`}{A list column containing a summary of the
values taken by the variable in Wave 1 of the panel.}
\item{`marginals_2`}{A list column containing a summary of the
values taken by the variable in Wave 2 of the panel.}
\item{`marginals_3`}{A list column containing a summary of the
values taken by the variable in Wave 3 of the panel.}
}
}
\source{
National Opinion Research Center, [http://gss.norc.org](http://gss.norc.org).
}
\usage{
data(gss_panel_doc)
}
\description{
A tibble containing information on the variable in the GSS
Three Wave Panel Data File, in wide format. See http://gss.norc.org/Get-Documentation for full
documentation of the variables.
}
\keyword{datasets}
