% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_panel_long}
\alias{gss_panel_long}
\title{General Social Survey Survey Three Wave Panel Data}
\format{An object of class \code{panel_data} (a tibble produced by
    the \code{panelr} package) with 6,132 rows and 1,199 columns.
\describe{
The GSS Codebook is the authoritative source for the variables in
}}
\source{
National Opinion Research Center, \href{http://gss.norc.org}{http://gss.norc.org}.
}
\usage{
data(gss_all)
}
\description{
A tibble containing the General Social Survey Thee Wave Panel Data
File, in long format. Variables are encoded as labelled vectors. See \href{http://gss.norc.org/Get-Documentation}{http://gss.norc.org/Get-Documentation} for full
documentation of the variables, and the `gss_doc` tibble contained in this package.
}
\keyword{datasets}
