% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_sub}
\alias{gss_sub}
\title{Example subset of the GSS Cumulative Data File 1972-2022}
\format{
A tibble with 72,390 rows and 19 columns.
\describe{
\item{\code{year}}{Year of the survey.}
\item{\code{id}}{Respondent id.}
\item{\code{ballot}}{Survey ballot}
\item{\code{age}}{Age of respondent}
\item{\code{race}}{Race of respondent}
\item{\code{sex}}{Sex of respondent}
\item{\code{degree}}{Highest level of education obtained}
\item{\code{padeg}}{Father's education}
\item{\code{padeg}}{Mother's education}
\item{\code{relig}}{Religion (simple coding)}
\item{\code{Polviews}}{Political views}
\item{\code{fefam}}{Response to a statement that it is better for man to go out to work, and for a woman to tend the home}
\item{\code{vpsu}}{Variance primary sampling unit}
\item{\code{vstrat}}{Variance stratum}
\item{\code{oversamp}}{Weights for black oversamples}
\item{\code{formwt}}{Survey weight for experimental randomization}
\item{\code{wtssall}}{Survey weight (1972-2018)}
\item{\code{wtssps}}{Poststratification survey weight (1972-2022)}
\item{\code{sampcode}}{Sampling error code}
\item{\code{sample}}{Sampling frame and method}
}
}
\source{
National Opinion Research Center, \url{http://gss.norc.org}.
}
\usage{
data(gss_sub)
}
\description{
A tibble containing just a few variables from the GSS
Cumulative Data File. See http://gss.norc.org/Get-Documentation for full
documentation of the variables.
}
\keyword{datasets}
