% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gssr-package.R
\name{gss_which_ballots}
\alias{gss_which_ballots}
\title{gss_which_ballots}
\usage{
gss_which_ballots(question)
}
\arguments{
\item{question}{A gss variable name (in lower case).}
}
\value{
A tibble with the name of the variable and the ballot information.
}
\description{
See which ballots (if any) a question appeared on across GSS years
}
\details{
What ballots did a particular question appear on in each GSS year?

Columns returned are
\itemize{
\item \code{variable} Variable name
\item \code{year} GSS year
\item \code{iap} Numbers in this column mean the question wasn't asked \emph{or} the ballot system was not in use.
\item \verb{ballot a}, \verb{ballot b}, etc: Count of the n responses on each ballot for a particular year.
}
}
\examples{
\dontrun{
gss_which_ballots(fefam)

}
}
