% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gssr-package.R
\name{gss_which_years}
\alias{gss_which_years}
\title{gss_which_years}
\usage{
gss_which_years(data, variable, year = year)
}
\arguments{
\item{data}{A tibble of data, usually \code{gss_all}}

\item{variable}{The variable or variables we want to check. Provide variables in tidyselect style, i.e. unquoted, and for multiple variables enclose unquoted in c()}

\item{year}{The grouping variable; defaults to and should always be \code{year}}
}
\value{
A tibble showing whether the question or questions were asked in each of the GSS years
}
\description{
See which years a particular question was asked in the GSS.
}
\details{
What years was a particular question asked in the GSS?
}
\examples{
\dontrun{
data(gss_all)
gss_all \%>\%
  gss_which_years(fefam)

gss_all \%>\%
  gss_which_years(c(industry, indus80, wrkgovt, commute))
}
}
