% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_591.R
\name{hefinfo}
\alias{hefinfo}
\title{Number of hef informant}
\source{
General Social Survey https://gss.norc.org
}
\description{
hefinfo
}
\details{
Question 1623. Number in family of person who was household informant (i.e. who provided information on household members).
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab not available in this year \tab 11th person \tab 1st person \tab 2nd person \tab 3rd person \tab 4th person \tab 5th person \tab 6th person \tab no answer \tab 7th person \tab 8th person \tab 9th person \tab 14th person \tab 12th person \tab 13th person \tab non-household person \tab 10th person \tab don't know \tab not available in this release \tab iap \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 5 \tab 769 \tab 596 \tab 56 \tab 11 \tab 2 \tab 2 \tab 27 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 6 \tab 1058 \tab 689 \tab 69 \tab 14 \tab 4 \tab 1 \tab 15 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 6 \tab 870 \tab 629 \tab 52 \tab 10 \tab 8 \tab 2 \tab 19 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 7 \tab 817 \tab 573 \tab 50 \tab 16 \tab 6 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 17 \tab 873 \tab 566 \tab 50 \tab 13 \tab 3 \tab - \tab 10 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 1 \tab 819 \tab 538 \tab 57 \tab 16 \tab 4 \tab 2 \tab 13 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 19 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 2 \tab 1086 \tab 602 \tab 64 \tab 25 \tab 4 \tab 4 \tab 16 \tab - \tab - \tab - \tab - \tab - \tab - \tab 16 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab - \tab 850 \tab 522 \tab 55 \tab 10 \tab 5 \tab 3 \tab 20 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 14 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 3 \tab 881 \tab 557 \tab 49 \tab 12 \tab 8 \tab 1 \tab 8 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 16 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 2 \tab 844 \tab 457 \tab 34 \tab 7 \tab 7 \tab 1 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab 13 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab - \tab 912 \tab 497 \tab 57 \tab 13 \tab 3 \tab 1 \tab 17 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 16 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 6 \tab 971 \tab 491 \tab 39 \tab 12 \tab 4 \tab 2 \tab 52 \tab - \tab - \tab - \tab - \tab - \tab - \tab 28 \tab 1 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 3 \tab 1929 \tab 909 \tab 77 \tab 22 \tab 9 \tab 4 \tab 18 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 20 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 2 \tab 1821 \tab 872 \tab 85 \tab 29 \tab 6 \tab - \tab 68 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 19 \tab - \tab 1 \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 2 \tab 1813 \tab 813 \tab 89 \tab 16 \tab 2 \tab 1 \tab 81 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 13 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 2 \tab 2124 \tab 529 \tab 64 \tab 19 \tab 5 \tab - \tab 52 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 18 \tab 2 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 1 \tab 2164 \tab 503 \tab 66 \tab 18 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 7 \tab - \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 3 \tab 2132 \tab 553 \tab 73 \tab 13 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 30 \tab - \tab 4 \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 6 \tab 3431 \tab 879 \tab 124 \tab 20 \tab 14 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 29 \tab - \tab 2 \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 2 \tab 1484 \tab 441 \tab 63 \tab 9 \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 13 \tab - \tab 3 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 5 \tab 1561 \tab 387 \tab 44 \tab 14 \tab 4 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 22 \tab - \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 5 \tab 1516 \tab 359 \tab 57 \tab 14 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 13 \tab - \tab 4 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 3 \tab 2026 \tab 419 \tab 53 \tab 13 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 22 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 6 \tab 2292 \tab 445 \tab 76 \tab 12 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 30 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 3 \tab 1902 \tab 265 \tab 130 \tab 26 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 16 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 10652 \tab 98 \tab 36945 \tab 14091 \tab 1633 \tab 384 \tab 120 \tab 33 \tab 434 \tab 9 \tab 6 \tab 4 \tab 1 \tab 9 \tab 1 \tab 374 \tab 3 \tab 17 \tab 4032 \tab 3544 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} 1st person
\item \code{2} 2nd person
\item \code{3} 3rd person
\item \code{4} 4th person
\item \code{5} 5th person
\item \code{6} 6th person
\item \code{7} 7th person
\item \code{8} 8th person
\item \code{9} 9th person
\item \code{10} 10th person
\item \code{11} 11th person
\item \code{12} 12th person
\item \code{13} 13th person
\item \code{14} 14th person
\item \code{22} non-household person
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
