% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_93.R
\name{helpblk}
\alias{helpblk}
\title{Should govt aid blacks?}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpblk
}
\details{
Question 312. Now look at card BF. Some people think that (blacks/negroes/African-Americans) have been discriminated against for so long that the government has a special obligation to help improve their living standards; they are at point 1. Others believe that the government should not be giving special treatment to (blacks/negroes/African-Americans); they are at point 5. a. Where would you place yourself on this scale, or haven't you made up your mind on this?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 2 \tab 4 \tab agree with both \tab don't know \tab government should help \tab no answer \tab no special treatment \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 126 \tab 179 \tab 311 \tab 40 \tab 240 \tab 2 \tab 592 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 151 \tab 322 \tab 406 \tab 66 \tab 120 \tab 7 \tab 527 \tab - \tab 1599 \cr
   1984 \tab - \tab 137 \tab 248 \tab 439 \tab 51 \tab 136 \tab 23 \tab 439 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 140 \tab 260 \tab 420 \tab 39 \tab 120 \tab 4 \tab 487 \tab - \tab 1470 \cr
   1987 \tab - \tab 204 \tab 279 \tab 519 \tab 46 \tab 278 \tab 8 \tab 485 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 99 \tab 172 \tab 285 \tab 30 \tab 74 \tab 4 \tab 333 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 97 \tab 190 \tab 278 \tab 36 \tab 90 \tab 2 \tab 342 \tab - \tab 1537 \cr
   1990 \tab - \tab 138 \tab 220 \tab 452 \tab 46 \tab 139 \tab 3 \tab 374 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 113 \tab 189 \tab 311 \tab 32 \tab 104 \tab 4 \tab 264 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 105 \tab 217 \tab 333 \tab 30 \tab 75 \tab 7 \tab 290 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 146 \tab 465 \tab 584 \tab 60 \tab 164 \tab 11 \tab 581 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 163 \tab 443 \tab 552 \tab 63 \tab 155 \tab 10 \tab 539 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 194 \tab 399 \tab 591 \tab 70 \tab 124 \tab 7 \tab 526 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 173 \tab 361 \tab 598 \tab 81 \tab 176 \tab 15 \tab 492 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 72 \tab 177 \tab 282 \tab 20 \tab 76 \tab 2 \tab 283 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 62 \tab 158 \tab 278 \tab 12 \tab 66 \tab 5 \tab 295 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 147 \tab 365 \tab 662 \tab 37 \tab 209 \tab 4 \tab 565 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 107 \tab 253 \tab 440 \tab 50 \tab 125 \tab 7 \tab 383 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 115 \tab 266 \tab 405 \tab 49 \tab 124 \tab 6 \tab 412 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 117 \tab 231 \tab 394 \tab 47 \tab 93 \tab 9 \tab 447 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 145 \tab 293 \tab 505 \tab 58 \tab 150 \tab 7 \tab 529 \tab - \tab 2538 \cr
   2016 \tab 913 \tab 220 \tab 371 \tab 608 \tab 72 \tab 246 \tab 7 \tab 430 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 188 \tab 253 \tab 507 \tab 61 \tab 235 \tab 9 \tab 310 \tab - \tab 2348 \cr
   2021 \tab 1382 \tab 407 \tab 315 \tab 679 \tab 11 \tab 560 \tab 3 \tab 641 \tab 34 \tab 4032 \cr
   2022 \tab 1174 \tab 287 \tab 256 \tab 707 \tab 60 \tab 498 \tab 13 \tab 535 \tab 14 \tab 3544 \cr
   Total \tab 33237 \tab 3853 \tab 6882 \tab 11546 \tab 1167 \tab 4377 \tab 179 \tab 11101 \tab 48 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} government should help
\item \code{3} agree with both
\item \code{5} no special treatment
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
