% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_50.R
\name{helpful}
\alias{helpful}
\title{helpful}
\source{
General Social Survey https://gss.norc.org
}
\description{
People helpful or looking out for selves
}
\details{
Question 161. Would you say that most of the time people try to be helpful, or that they are mostly just looking out for themselves?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab depends \tab don't know \tab iap \tab looking out for themselves \tab no answer \tab try to be helpful \tab not available in this year \tab Total \cr
   1972 \tab 100 \tab 22 \tab 1 \tab 740 \tab 1 \tab 749 \tab - \tab 1613 \cr
   1973 \tab 56 \tab 4 \tab - \tab 739 \tab 3 \tab 702 \tab - \tab 1504 \cr
   1974 \tab - \tab - \tab 1484 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 99 \tab 8 \tab - \tab 545 \tab 2 \tab 836 \tab - \tab 1490 \cr
   1976 \tab 91 \tab 5 \tab - \tab 756 \tab 1 \tab 646 \tab - \tab 1499 \cr
   1977 \tab - \tab - \tab 1530 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 80 \tab 4 \tab - \tab 536 \tab 5 \tab 907 \tab - \tab 1532 \cr
   1980 \tab 64 \tab 9 \tab - \tab 678 \tab 3 \tab 714 \tab - \tab 1468 \cr
   1982 \tab - \tab - \tab 1860 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 68 \tab 11 \tab - \tab 603 \tab - \tab 917 \tab - \tab 1599 \cr
   1984 \tab 54 \tab 3 \tab - \tab 650 \tab 3 \tab 763 \tab - \tab 1473 \cr
   1985 \tab - \tab - \tab 1534 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 76 \tab 7 \tab - \tab 565 \tab 5 \tab 817 \tab - \tab 1470 \cr
   1987 \tab 81 \tab 11 \tab - \tab 935 \tab 5 \tab 787 \tab - \tab 1819 \cr
   1988 \tab 46 \tab 6 \tab 484 \tab 448 \tab 3 \tab 494 \tab - \tab 1481 \cr
   1989 \tab 52 \tab 5 \tab 502 \tab 459 \tab 1 \tab 518 \tab - \tab 1537 \cr
   1990 \tab 55 \tab 7 \tab 473 \tab 372 \tab 1 \tab 464 \tab - \tab 1372 \cr
   1991 \tab 72 \tab 8 \tab 500 \tab 436 \tab 5 \tab 496 \tab - \tab 1517 \cr
   1993 \tab 60 \tab 14 \tab 549 \tab 431 \tab 1 \tab 551 \tab - \tab 1606 \cr
   1994 \tab 128 \tab 9 \tab 981 \tab 935 \tab 7 \tab 932 \tab - \tab 2992 \cr
   1996 \tab 138 \tab 10 \tab 979 \tab 940 \tab 4 \tab 833 \tab - \tab 2904 \cr
   1998 \tab 176 \tab 7 \tab 921 \tab 815 \tab 2 \tab 911 \tab - \tab 2832 \cr
   2000 \tab 166 \tab 10 \tab 921 \tab 851 \tab 3 \tab 866 \tab - \tab 2817 \cr
   2002 \tab 88 \tab 2 \tab 1853 \tab 391 \tab - \tab 431 \tab - \tab 2765 \cr
   2004 \tab 76 \tab 3 \tab 1936 \tab 359 \tab - \tab 438 \tab - \tab 2812 \cr
   2006 \tab 194 \tab 4 \tab 2521 \tab 924 \tab - \tab 867 \tab - \tab 4510 \cr
   2008 \tab 118 \tab 2 \tab 658 \tab 624 \tab - \tab 621 \tab - \tab 2023 \cr
   2010 \tab 135 \tab 5 \tab 667 \tab 588 \tab 2 \tab 647 \tab - \tab 2044 \cr
   2012 \tab 98 \tab 8 \tab 636 \tab 612 \tab 2 \tab 618 \tab - \tab 1974 \cr
   2014 \tab 145 \tab 5 \tab 851 \tab 742 \tab - \tab 795 \tab - \tab 2538 \cr
   2016 \tab 137 \tab 7 \tab 911 \tab 892 \tab 3 \tab 917 \tab - \tab 2867 \cr
   2018 \tab 142 \tab 9 \tab 785 \tab 694 \tab 1 \tab 717 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 149 \tab 9 \tab 2325 \tab 549 \tab - \tab 512 \tab - \tab 3544 \cr
   Total \tab 2944 \tab 214 \tab 25862 \tab 18809 \tab 63 \tab 20466 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} try to be helpful
\item \code{2} looking out for themselves
\item \code{3} depends
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
