% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_93.R
\name{helpnot}
\alias{helpnot}
\title{Should govt do more or less?}
\source{
General Social Survey https://gss.norc.org
}
\description{
helpnot
}
\details{
Question 310. Please look at card BD. Some people think that the government in Washington is trying to do too many things that should be left to individuals and private busienss; they are at point 5 on this card. Others disagree and think that the government should do even more to solve our country's problems; they are at point 1. a. Where would you place yourself on this scale, or haven't you made up your mind on this?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 2 \tab 4 \tab agree with both \tab don't know \tab government does too much \tab government should do more \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 161 \tab 173 \tab 433 \tab 104 \tab 245 \tab 371 \tab 3 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 175 \tab 280 \tab 574 \tab 110 \tab 258 \tab 198 \tab 4 \tab - \tab 1599 \cr
   1984 \tab - \tab 184 \tab 241 \tab 537 \tab 79 \tab 204 \tab 204 \tab 24 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 185 \tab 228 \tab 604 \tab 81 \tab 189 \tab 180 \tab 3 \tab - \tab 1470 \cr
   1987 \tab - \tab 260 \tab 241 \tab 657 \tab 104 \tab 220 \tab 328 \tab 9 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 138 \tab 141 \tab 392 \tab 60 \tab 128 \tab 137 \tab 1 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 147 \tab 159 \tab 390 \tab 72 \tab 131 \tab 134 \tab 2 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 135 \tab 133 \tab 364 \tab 54 \tab 95 \tab 118 \tab - \tab - \tab 1372 \cr
   1991 \tab 500 \tab 149 \tab 148 \tab 395 \tab 70 \tab 111 \tab 139 \tab 5 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 156 \tab 162 \tab 401 \tab 39 \tab 155 \tab 138 \tab 6 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 271 \tab 358 \tab 709 \tab 98 \tab 325 \tab 235 \tab 15 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 239 \tab 324 \tab 722 \tab 100 \tab 304 \tab 229 \tab 7 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 231 \tab 339 \tab 749 \tab 105 \tab 282 \tab 202 \tab 3 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 226 \tab 331 \tab 709 \tab 95 \tab 313 \tab 213 \tab 9 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 93 \tab 123 \tab 431 \tab 27 \tab 108 \tab 129 \tab 1 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 86 \tab 123 \tab 381 \tab 22 \tab 126 \tab 137 \tab 1 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 199 \tab 288 \tab 859 \tab 51 \tab 250 \tab 339 \tab 3 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 174 \tab 175 \tab 539 \tab 53 \tab 194 \tab 226 \tab 4 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 165 \tab 189 \tab 498 \tab 58 \tab 257 \tab 205 \tab 5 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 147 \tab 171 \tab 530 \tab 45 \tab 242 \tab 199 \tab 4 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 174 \tab 258 \tab 624 \tab 65 \tab 342 \tab 223 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 913 \tab 210 \tab 273 \tab 754 \tab 92 \tab 292 \tab 329 \tab 4 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 219 \tab 207 \tab 644 \tab 68 \tab 190 \tab 231 \tab 4 \tab - \tab 2348 \cr
   2021 \tab 1377 \tab 386 \tab 319 \tab 977 \tab 10 \tab 388 \tab 539 \tab - \tab 36 \tab 4032 \cr
   2022 \tab 1174 \tab 260 \tab 255 \tab 906 \tab 71 \tab 335 \tab 520 \tab 7 \tab 16 \tab 3544 \cr
   Total \tab 33705 \tab 4770 \tab 5639 \tab 14779 \tab 1733 \tab 5684 \tab 5903 \tab 125 \tab 52 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} government should do more
\item \code{3} agree with both
\item \code{5} government does too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
