% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_58.R
\name{helpoth}
\alias{helpoth}
\title{helpoth}
\source{
General Social Survey https://gss.norc.org
}
\description{
To help others
}
\details{
Question 170. If you had to choose, which thing on this list would you pick as the most important for a child to learn to prepare him or her for life? a. Which comes next in importantc? b. Which comes third? c. Which comes fourth? TO HELP OTHERS WHEN THEY NEED HELP
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab fifth \tab first \tab fourth \tab no answer \tab second \tab third \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 730 \tab 30 \tab 100 \tab 143 \tab 8 \tab 216 \tab 243 \tab 1470 \cr
   1987 \tab - \tab 74 \tab 209 \tab 371 \tab 19 \tab 516 \tab 630 \tab 1819 \cr
   1988 \tab 484 \tab 44 \tab 127 \tab 206 \tab 3 \tab 332 \tab 285 \tab 1481 \cr
   1989 \tab 502 \tab 40 \tab 123 \tab 196 \tab 35 \tab 312 \tab 329 \tab 1537 \cr
   1990 \tab 473 \tab 29 \tab 119 \tab 179 \tab 28 \tab 273 \tab 271 \tab 1372 \cr
   1991 \tab 500 \tab 33 \tab 126 \tab 175 \tab 35 \tab 316 \tab 332 \tab 1517 \cr
   1993 \tab 549 \tab 37 \tab 135 \tab 208 \tab 26 \tab 337 \tab 314 \tab 1606 \cr
   1994 \tab 981 \tab 61 \tab 252 \tab 381 \tab 33 \tab 567 \tab 717 \tab 2992 \cr
   1996 \tab 979 \tab 50 \tab 241 \tab 373 \tab 30 \tab 589 \tab 642 \tab 2904 \cr
   1998 \tab 921 \tab 49 \tab 253 \tab 375 \tab 31 \tab 567 \tab 636 \tab 2832 \cr
   2000 \tab 921 \tab 52 \tab 247 \tab 344 \tab 33 \tab 590 \tab 630 \tab 2817 \cr
   2002 \tab 1853 \tab 23 \tab 127 \tab 154 \tab 13 \tab 279 \tab 316 \tab 2765 \cr
   2004 \tab 1936 \tab 21 \tab 151 \tab 126 \tab 8 \tab 281 \tab 289 \tab 2812 \cr
   2006 \tab 3508 \tab 23 \tab 159 \tab 170 \tab 7 \tab 304 \tab 339 \tab 4510 \cr
   2008 \tab 658 \tab 24 \tab 217 \tab 188 \tab 14 \tab 443 \tab 479 \tab 2023 \cr
   2010 \tab 667 \tab 19 \tab 237 \tab 211 \tab 28 \tab 440 \tab 442 \tab 2044 \cr
   2012 \tab 636 \tab 19 \tab 214 \tab 207 \tab 15 \tab 452 \tab 431 \tab 1974 \cr
   2014 \tab 851 \tab 26 \tab 311 \tab 207 \tab 17 \tab 528 \tab 598 \tab 2538 \cr
   2016 \tab 911 \tab 23 \tab 357 \tab 230 \tab 22 \tab 656 \tab 668 \tab 2867 \cr
   2018 \tab 785 \tab 16 \tab 312 \tab 173 \tab 21 \tab 533 \tab 508 \tab 2348 \cr
   2021 \tab 1360 \tab 22 \tab 533 \tab 214 \tab 99 \tab 878 \tab 926 \tab 4032 \cr
   2022 \tab 1173 \tab 31 \tab 470 \tab 181 \tab 77 \tab 768 \tab 844 \tab 3544 \cr
   Total \tab 39964 \tab 746 \tab 5020 \tab 5012 \tab 602 \tab 10177 \tab 10869 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} first
\item \code{2} second
\item \code{3} third
\item \code{4} fourth
\item \code{5} fifth
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
