% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_93.R
\name{helppoor}
\alias{helppoor}
\title{Should govt improve standard of living?}
\source{
General Social Survey https://gss.norc.org
}
\description{
helppoor
}
\details{
Question 309. I'd like to talk with you about issues some people tell us are important. Please look at card BC. Some people think that the government in Washington should do everything possible to improve the standard of living of all poor Americans; they are at point 1 on this card. Other people think itis not the government's responsibility, and that each person should take care of himself; they are at point 5. a. Where would you place yourself on this scale, or haven't you made up your mind on this?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 2 \tab 4 \tab agree with both \tab don't know \tab government should help \tab people should care for themselves \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 147 \tab 155 \tab 518 \tab 42 \tab 434 \tab 194 \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 236 \tab 215 \tab 631 \tab 63 \tab 268 \tab 180 \tab 6 \tab - \tab 1599 \cr
   1984 \tab - \tab 160 \tab 207 \tab 666 \tab 53 \tab 253 \tab 123 \tab 11 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 181 \tab 168 \tab 657 \tab 36 \tab 267 \tab 157 \tab 4 \tab - \tab 1470 \cr
   1987 \tab - \tab 220 \tab 215 \tab 757 \tab 46 \tab 395 \tab 175 \tab 11 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 127 \tab 118 \tab 441 \tab 24 \tab 171 \tab 114 \tab 2 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 154 \tab 141 \tab 445 \tab 34 \tab 170 \tab 89 \tab 2 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 134 \tab 111 \tab 387 \tab 23 \tab 168 \tab 75 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 164 \tab 132 \tab 435 \tab 29 \tab 171 \tab 82 \tab 4 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 142 \tab 145 \tab 501 \tab 25 \tab 131 \tab 109 \tab 4 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 268 \tab 318 \tab 870 \tab 58 \tab 252 \tab 232 \tab 13 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 231 \tab 297 \tab 871 \tab 67 \tab 248 \tab 204 \tab 7 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 240 \tab 316 \tab 818 \tab 62 \tab 239 \tab 230 \tab 6 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 237 \tab 309 \tab 786 \tab 65 \tab 263 \tab 227 \tab 9 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 95 \tab 115 \tab 420 \tab 23 \tab 154 \tab 104 \tab 1 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 87 \tab 124 \tab 410 \tab 11 \tab 148 \tab 96 \tab - \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 204 \tab 261 \tab 915 \tab 30 \tab 369 \tab 209 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 168 \tab 187 \tab 568 \tab 31 \tab 267 \tab 143 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 151 \tab 198 \tab 597 \tab 32 \tab 226 \tab 168 \tab 5 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 145 \tab 181 \tab 580 \tab 40 \tab 212 \tab 176 \tab 4 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 210 \tab 241 \tab 695 \tab 48 \tab 273 \tab 219 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 913 \tab 281 \tab 287 \tab 802 \tab 61 \tab 345 \tab 172 \tab 6 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 220 \tab 186 \tab 692 \tab 44 \tab 261 \tab 160 \tab - \tab - \tab 2348 \cr
   2021 \tab 1373 \tab 335 \tab 327 \tab 1035 \tab 10 \tab 675 \tab 261 \tab - \tab 16 \tab 4032 \cr
   2022 \tab 1174 \tab 270 \tab 236 \tab 922 \tab 34 \tab 652 \tab 242 \tab 7 \tab 7 \tab 3544 \cr
   Total \tab 33701 \tab 4807 \tab 5190 \tab 16419 \tab 991 \tab 7012 \tab 4141 \tab 106 \tab 23 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} government should help
\item \code{3} agree with both
\item \code{5} people should care for themselves
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
