% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_93.R
\name{helpsick}
\alias{helpsick}
\title{helpsick}
\source{
General Social Survey https://gss.norc.org
}
\description{
Should govt help pay for medical care?
}
\details{
Question 311. Look at card BE. In general, some people think that it is the responsibility of the government in Washington to see to it that people have help in paying for doctors and hospital bills; they are at point 1. Others think that these matters are not the responsibility of the federal government and that people should take care of these things themselves; they are at point 5. a. Where would you place yourself on this scale, or haven't you made up your mind on this?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 2 \tab 4 \tab agree with both \tab don't know \tab government should help \tab no answer \tab people should care for themselves \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 188 \tab 115 \tab 423 \tab 31 \tab 532 \tab 6 \tab 195 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 300 \tab 162 \tab 504 \tab 53 \tab 413 \tab 6 \tab 161 \tab - \tab 1599 \cr
   1984 \tab - \tab 269 \tab 176 \tab 506 \tab 40 \tab 347 \tab 24 \tab 111 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 292 \tab 164 \tab 462 \tab 36 \tab 416 \tab 7 \tab 93 \tab - \tab 1470 \cr
   1987 \tab - \tab 353 \tab 141 \tab 610 \tab 39 \tab 549 \tab 6 \tab 121 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 214 \tab 85 \tab 345 \tab 21 \tab 264 \tab 4 \tab 64 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 231 \tab 78 \tab 310 \tab 21 \tab 314 \tab 4 \tab 77 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 229 \tab 67 \tab 270 \tab 26 \tab 266 \tab 2 \tab 39 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 250 \tab 93 \tab 271 \tab 20 \tab 321 \tab 4 \tab 58 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 234 \tab 98 \tab 332 \tab 23 \tab 297 \tab 8 \tab 65 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 420 \tab 217 \tab 620 \tab 53 \tab 503 \tab 13 \tab 185 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 403 \tab 198 \tab 622 \tab 51 \tab 518 \tab 7 \tab 126 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 423 \tab 185 \tab 610 \tab 56 \tab 473 \tab 6 \tab 158 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 402 \tab 194 \tab 586 \tab 57 \tab 539 \tab 9 \tab 109 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 175 \tab 65 \tab 308 \tab 13 \tab 287 \tab 1 \tab 63 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 175 \tab 66 \tab 274 \tab 11 \tab 280 \tab 1 \tab 69 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 367 \tab 166 \tab 634 \tab 25 \tab 665 \tab - \tab 132 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 252 \tab 125 \tab 384 \tab 32 \tab 471 \tab 2 \tab 99 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 224 \tab 142 \tab 422 \tab 45 \tab 402 \tab 6 \tab 136 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 228 \tab 153 \tab 410 \tab 39 \tab 384 \tab 2 \tab 122 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 311 \tab 175 \tab 533 \tab 46 \tab 412 \tab 1 \tab 209 \tab - \tab 2538 \cr
   2016 \tab 913 \tab 364 \tab 190 \tab 599 \tab 55 \tab 587 \tab 3 \tab 156 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 306 \tab 129 \tab 448 \tab 33 \tab 528 \tab 2 \tab 117 \tab - \tab 2348 \cr
   2021 \tab 1377 \tab 471 \tab 227 \tab 793 \tab 4 \tab 917 \tab 1 \tab 219 \tab 23 \tab 4032 \cr
   2022 \tab 1174 \tab 381 \tab 140 \tab 687 \tab 33 \tab 949 \tab 4 \tab 166 \tab 10 \tab 3544 \cr
   Total \tab 33705 \tab 7462 \tab 3551 \tab 11963 \tab 863 \tab 11634 \tab 129 \tab 3050 \tab 33 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} government should help
\item \code{3} agree with both
\item \code{5} people should care for themselves
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
