% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_591.R
\name{hhtype}
\alias{hhtype}
\title{Household type}
\source{
General Social Survey https://gss.norc.org
}
\description{
hhtype
}
\details{
Question 1626. Household Type.
}
\section{Values}{

\itemize{
\item \code{1} 1adlt,0kids
\item \code{2} 1adlt,1+kids
\item \code{3} 2adlts,mar,0kids
\item \code{4} 2adlts,mar,1+kids
\item \code{5} 2adlts,dkmar,0kids
\item \code{6} 2adlts,dkmar,1+kids
\item \code{7} 2as,ntmar,ntrel,opsex,0ks
\item \code{8} 2as,ntmar,ntrel,opsex,1+ks
\item \code{9} 2as,ntmar,ntrel,smsex,0ks
\item \code{10} 2as,ntmar,ntrel,smsex,1+ks
\item \code{11} 2adlts,ntmar,rel,0kids
\item \code{12} 2adlts,ntmar,rel,1+kids
\item \code{13} 2adlts,ntmar,dkrel,0kids
\item \code{14} 2adlts,ntmar,dkrel,1+kids
\item \code{15} 2as,ntmar,ntrel,dksex,0ks
\item \code{16} 2as,ntmar,ntrel,dksex,1+ks
\item \code{20} 3adlt,0mar,sp,0kids
\item \code{21} 3adlts,0mar,sp,rel,0kids
\item \code{22} 3adlts,0mar,rel,sp,0kids
\item \code{23} 3adlts,0mar,rel,0kids
\item \code{24} 3adlts,0mar,sp,ntrel,0kids
\item \code{25} 3adlts,0mar,ntrel,sp,0kids
\item \code{26} 3adlts,0mar,rel,0kids
\item \code{27} 3adlts,0mar,ntrel,rel,0kids
\item \code{28} 3adlts,0mar,ntrel,0kids
\item \code{30} 3adlts,1mar,sp,0kids
\item \code{31} 3adlts,1mar,sp,rel,0kids
\item \code{32} 3adlts,1mar,rel,sp,0kids
\item \code{33} 3adlts,1mar,rel,rel,0kids
\item \code{34} 3adlts,1mar,sp,ntrel,0kids
\item \code{35} 3adlts,1mar,ntrel,sp,0kids
\item \code{36} 3adlts,1mar,rel,ntrel,0kids
\item \code{37} 3adlts,1mar,ntrel,rel,0kids
\item \code{38} 3adlts,1mar,ntrel,0kids
\item \code{40} 3adlts,2mar,sp,0kids
\item \code{41} 3adlts,2mar,sp,rel,0kids
\item \code{42} 3adlts,2mar,rel,sp,0kids
\item \code{43} 3adlts,2mar,rel,0kids
\item \code{44} 3adlts,2mar,sp,ntrel,0kids
\item \code{45} 3adlts,2mar,rel,ntrel,0kids
\item \code{46} 3adlts,2mar,rel,ntrel,0kids
\item \code{47} 3adlts,2mar,ntrel,rel,0kids
\item \code{48} 3adlts,2mar,ntrel,0kids
\item \code{50} 3adlts,3mar,sp,0kids
\item \code{51} 3adlts,3mar,sp,rel,0kids
\item \code{52} 3adlts,3mar,sp,rel,0kids
\item \code{53} 3adlts,3mar,rel,0kids
\item \code{54} 3adlts,3mar,sp,ntrel,0kids
\item \code{55} 3adlts,3mar,ntrel,sp,0kids
\item \code{56} 3adlts,3mar,rel,ntrel,0kids
\item \code{57} 3adlts,3mar,ntrel,rel,0kids
\item \code{58} 3adlts,3mar,ntrel,0kids
\item \code{100} 4+adlts,0mar,0kids
\item \code{101} 4+adlts,1mar,0kids
\item \code{102} 4+adlts,2mar,0kids
\item \code{103} 4+adlts,3mar,0kids
\item \code{104} 4+adlts,4+mar,0kids
\item \code{120} 3adlt,0mar,sp,1+kids
\item \code{121} 3adlts,0mar,sp,rel,1+kids
\item \code{122} 3adlts,0mar,rel,sp,1+kids
\item \code{123} 3adlts,0mar,rel,1+kids
\item \code{124} 3adlts,0mar,sp,ntrel,1+kids
\item \code{125} 3adlts,0mar,ntrel,sp,1+kids
\item \code{126} 3adlts,0mar,rel,1+kids
\item \code{127} 3adlts,0mar,ntrel,rel,1+kids
\item \code{128} 3adlts,0mar,ntrel,1+kids
\item \code{130} 3adlts,1mar,sp,1+kids
\item \code{131} 3adlts,1mar,sp,rel,1+kids
\item \code{132} 3adlts,1mar,rel,sp,1+kids
\item \code{133} 3adlts,1mar,rel,rel,1+kids
\item \code{134} 3adlts,1mar,sp,ntrel,1+kids
\item \code{135} 3adlts,1mar,ntrel,sp,1+kids
\item \code{136} 3adlts,1mar,rel,ntrel,1+kids
\item \code{137} 3adlts,1mar,ntrel,rel,1+kids
\item \code{138} 3adlts,1mar,ntrel,1+kids
\item \code{140} 3adlts,2mar,sp,1+kids
\item \code{141} 3adlts,2mar,sp,rel,1+kids
\item \code{142} 3adlts,2mar,rel,sp,1+kids
\item \code{143} 3adlts,2mar,rel,1+kids
\item \code{144} 3adlts,2mar,sp,ntrel,1+kids
\item \code{145} 3adlts,2mar,rel,ntrel,1+kids
\item \code{146} 3adlts,2mar,rel,ntrel,1+kids
\item \code{147} 3adlts,2mar,ntrel,rel,1+kids
\item \code{148} 3adlts,2mar,ntrel,1+kids
\item \code{150} 3adlts,3mar,sp,1+kids
\item \code{151} 3adlts,3mar,sp,rel,1+kids
\item \code{152} 3adlts,3mar,sp,rel,1+kids
\item \code{153} 3adlts,3mar,rel,1+kids
\item \code{154} 3adlts,3mar,sp,ntrel,1+kids
\item \code{155} 3adlts,3mar,ntrel,sp,1+kids
\item \code{156} 3adlts,3mar,rel,ntrel,1+kids
\item \code{157} 3adlts,3mar,ntrel,rel,1+kids
\item \code{158} 3adlts,3mar,ntrel,1+kids
\item \code{200} 4+adlts,0mar,1+kids
\item \code{201} 4+adlts,1mar,1+kids
\item \code{202} 4+adlts,2mar,1+kids
\item \code{203} 4+adlts,3mar,1+kids
\item \code{204} 4+adlts,4+mar,1+kids
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
