% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_591.R
\name{hhtype1}
\alias{hhtype1}
\title{hhtype1}
\source{
General Social Survey https://gss.norc.org
}
\description{
Household type (condensed)
}
\details{
Question 1627. Household type (condensed).
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab cohabitating couple with children \tab cohabitating couple, no children \tab married couple with children \tab married couple, no children \tab no answer \tab non-family with children \tab non-family, no children \tab other family with children \tab other family, no children \tab single adult \tab single parent \tab unsure with children \tab unsure, no children \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 10 \tab 14 \tab 610 \tab 471 \tab 10 \tab 3 \tab 16 \tab 28 \tab 56 \tab 200 \tab 60 \tab 4 \tab 8 \tab - \tab 1490 \cr
   1976 \tab 755 \tab - \tab 7 \tab 274 \tab 227 \tab - \tab 1 \tab 10 \tab 12 \tab 23 \tab 138 \tab 45 \tab 2 \tab 5 \tab - \tab 1499 \cr
   1977 \tab - \tab 2 \tab 12 \tab 564 \tab 480 \tab 3 \tab 1 \tab 23 \tab 33 \tab 46 \tab 263 \tab 85 \tab 6 \tab 12 \tab - \tab 1530 \cr
   1978 \tab - \tab 5 \tab 13 \tab 573 \tab 437 \tab 5 \tab 1 \tab 26 \tab 38 \tab 49 \tab 297 \tab 79 \tab 5 \tab 4 \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab 18 \tab 481 \tab 448 \tab 7 \tab 4 \tab 26 \tab 33 \tab 53 \tab 287 \tab 85 \tab 13 \tab 13 \tab - \tab 1468 \cr
   1982 \tab - \tab 9 \tab 30 \tab 540 \tab 528 \tab 5 \tab 2 \tab 29 \tab 62 \tab 74 \tab 410 \tab 134 \tab 18 \tab 19 \tab - \tab 1860 \cr
   1983 \tab - \tab 5 \tab 20 \tab 540 \tab 482 \tab 3 \tab 4 \tab 21 \tab 38 \tab 56 \tab 309 \tab 95 \tab 16 \tab 10 \tab - \tab 1599 \cr
   1984 \tab - \tab 13 \tab 18 \tab 446 \tab 433 \tab 1 \tab 2 \tab 39 \tab 29 \tab 46 \tab 330 \tab 91 \tab 8 \tab 17 \tab - \tab 1473 \cr
   1985 \tab - \tab 8 \tab 20 \tab 459 \tab 462 \tab 5 \tab 2 \tab 49 \tab 22 \tab 55 \tab 342 \tab 91 \tab 8 \tab 11 \tab - \tab 1534 \cr
   1986 \tab - \tab 10 \tab 18 \tab 459 \tab 421 \tab 5 \tab 2 \tab 36 \tab 33 \tab 61 \tab 313 \tab 71 \tab 13 \tab 28 \tab - \tab 1470 \cr
   1987 \tab - \tab 12 \tab 30 \tab 502 \tab 465 \tab 15 \tab 4 \tab 52 \tab 56 \tab 92 \tab 420 \tab 109 \tab 30 \tab 32 \tab - \tab 1819 \cr
   1988 \tab - \tab 11 \tab 31 \tab 407 \tab 425 \tab 5 \tab 1 \tab 49 \tab 38 \tab 52 \tab 326 \tab 115 \tab 8 \tab 13 \tab - \tab 1481 \cr
   1989 \tab - \tab 14 \tab 32 \tab 457 \tab 440 \tab 12 \tab 1 \tab 45 \tab 29 \tab 55 \tab 327 \tab 93 \tab 6 \tab 26 \tab - \tab 1537 \cr
   1990 \tab - \tab 11 \tab 36 \tab 360 \tab 407 \tab 5 \tab 2 \tab 28 \tab 30 \tab 67 \tab 329 \tab 79 \tab 1 \tab 17 \tab - \tab 1372 \cr
   1991 \tab - \tab 12 \tab 35 \tab 404 \tab 434 \tab 6 \tab 3 \tab 35 \tab 31 \tab 53 \tab 375 \tab 103 \tab 8 \tab 18 \tab - \tab 1517 \cr
   1993 \tab - \tab 23 \tab 43 \tab 439 \tab 469 \tab 2 \tab 5 \tab 33 \tab 27 \tab 70 \tab 376 \tab 97 \tab 8 \tab 14 \tab - \tab 1606 \cr
   1994 \tab - \tab 43 \tab 74 \tab 777 \tab 845 \tab 6 \tab 8 \tab 58 \tab 54 \tab 119 \tab 759 \tab 211 \tab 13 \tab 25 \tab - \tab 2992 \cr
   1996 \tab - \tab 50 \tab 74 \tab 687 \tab 748 \tab 26 \tab 9 \tab 75 \tab 64 \tab 101 \tab 725 \tab 238 \tab 26 \tab 81 \tab - \tab 2904 \cr
   1998 \tab - \tab 46 \tab 64 \tab 630 \tab 754 \tab 16 \tab 7 \tab 74 \tab 61 \tab 107 \tab 771 \tab 181 \tab 40 \tab 81 \tab - \tab 2832 \cr
   2000 \tab - \tab 57 \tab 100 \tab 677 \tab 652 \tab 16 \tab 12 \tab 80 \tab 71 \tab 139 \tab 740 \tab 219 \tab 21 \tab 33 \tab - \tab 2817 \cr
   2002 \tab - \tab 61 \tab 85 \tab 471 \tab 784 \tab 8 \tab 9 \tab 65 \tab 60 \tab 139 \tab 870 \tab 175 \tab 7 \tab 31 \tab - \tab 2765 \cr
   2004 \tab - \tab 46 \tab 89 \tab 590 \tab 939 \tab 9 \tab 4 \tab 49 \tab 38 \tab 121 \tab 706 \tab 169 \tab 20 \tab 32 \tab - \tab 2812 \cr
   2006 \tab - \tab 88 \tab 139 \tab 970 \tab 1309 \tab 20 \tab 9 \tab 101 \tab 98 \tab 198 \tab 1220 \tab 270 \tab 43 \tab 45 \tab - \tab 4510 \cr
   2008 \tab - \tab 43 \tab 58 \tab 424 \tab 577 \tab 9 \tab 6 \tab 44 \tab 60 \tab 93 \tab 520 \tab 148 \tab 16 \tab 25 \tab - \tab 2023 \cr
   2010 \tab - \tab 40 \tab 79 \tab 342 \tab 582 \tab 14 \tab 8 \tab 21 \tab 58 \tab 115 \tab 601 \tab 127 \tab 18 \tab 39 \tab - \tab 2044 \cr
   2012 \tab - \tab 49 \tab 80 \tab 369 \tab 594 \tab 8 \tab 1 \tab 10 \tab 45 \tab 142 \tab 529 \tab 105 \tab 18 \tab 24 \tab - \tab 1974 \cr
   2014 \tab - \tab 47 \tab 98 \tab 379 \tab 839 \tab 13 \tab 2 \tab 17 \tab 46 \tab 199 \tab 695 \tab 133 \tab 21 \tab 49 \tab - \tab 2538 \cr
   2016 \tab - \tab 66 \tab 130 \tab 440 \tab 826 \tab 13 \tab 1 \tab 17 \tab 61 \tab 249 \tab 850 \tab 144 \tab 24 \tab 46 \tab - \tab 2867 \cr
   2018 \tab - \tab 42 \tab 120 \tab 343 \tab 705 \tab 17 \tab 7 \tab 55 \tab 65 \tab 135 \tab 669 \tab 106 \tab 23 \tab 61 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 8900 \tab 823 \tab 1567 \tab 14614 \tab 17183 \tab 264 \tab 121 \tab 1183 \tab 1320 \tab 2765 \tab 14697 \tab 3658 \tab 444 \tab 819 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} married couple, no children
\item \code{2} single parent
\item \code{3} other family, no children
\item \code{4} single adult
\item \code{5} cohabitating couple, no children
\item \code{6} non-family, no children
\item \code{8} unsure, no children
\item \code{11} married couple with children
\item \code{13} other family with children
\item \code{15} cohabitating couple with children
\item \code{16} non-family with children
\item \code{18} unsure with children
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
