% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_74.R
\name{hitage}
\alias{hitage}
\title{Beaten as child or adult}
\source{
General Social Survey https://gss.norc.org
}
\description{
hitage
}
\details{
Question 229a. If yes: Did this happen to you as a child or as an adult?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab adult \tab both \tab child \tab no answer \tab don't know \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1091 \tab 124 \tab 109 \tab 179 \tab 1 \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1017 \tab 155 \tab 99 \tab 214 \tab 3 \tab 2 \tab 1490 \cr
   1976 \tab 1075 \tab 151 \tab 96 \tab 176 \tab - \tab 1 \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 995 \tab 178 \tab 122 \tab 237 \tab - \tab - \tab 1532 \cr
   1980 \tab 977 \tab 183 \tab 98 \tab 208 \tab 2 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 869 \tab 237 \tab 179 \tab 311 \tab 3 \tab - \tab 1599 \cr
   1984 \tab 886 \tab 194 \tab 131 \tab 259 \tab 2 \tab 1 \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 935 \tab 197 \tab 123 \tab 212 \tab 3 \tab - \tab 1470 \cr
   1987 \tab 1161 \tab 239 \tab 139 \tab 268 \tab 9 \tab 3 \tab 1819 \cr
   1988 \tab 1129 \tab 139 \tab 67 \tab 143 \tab 3 \tab - \tab 1481 \cr
   1989 \tab 1159 \tab 135 \tab 81 \tab 161 \tab 1 \tab - \tab 1537 \cr
   1990 \tab 1042 \tab 123 \tab 74 \tab 132 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 1165 \tab 138 \tab 72 \tab 135 \tab 7 \tab - \tab 1517 \cr
   1993 \tab 1226 \tab 149 \tab 99 \tab 128 \tab 3 \tab 1 \tab 1606 \cr
   1994 \tab 2796 \tab 60 \tab 49 \tab 86 \tab 1 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 65554 \tab 2402 \tab 1538 \tab 2849 \tab 39 \tab 8 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} child
\item \code{2} adult
\item \code{3} both
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
