% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_70.R
\name{homosex}
\alias{homosex}
\title{Homosexual sex relations}
\source{
General Social Survey https://gss.norc.org
}
\description{
homosex
}
\details{
Question 219. What about sexual relations between two adults of the same sex?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab almost always wrong \tab always wrong \tab don't know \tab no answer \tab not wrong at all \tab other \tab wrong only sometimes \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 95 \tab 1053 \tab 49 \tab 7 \tab 159 \tab 31 \tab 110 \tab - \tab 1504 \cr
   1974 \tab - \tab 71 \tab 995 \tab 72 \tab - \tab 183 \tab 51 \tab 112 \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 88 \tab 999 \tab 62 \tab 11 \tab 227 \tab - \tab 112 \tab - \tab 1499 \cr
   1977 \tab - \tab 84 \tab 1044 \tab 69 \tab 8 \tab 216 \tab - \tab 109 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 84 \tab 1024 \tab 68 \tab 3 \tab 204 \tab - \tab 85 \tab - \tab 1468 \cr
   1982 \tab - \tab 93 \tab 1324 \tab 76 \tab 13 \tab 238 \tab - \tab 116 \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 70 \tab 1035 \tab 54 \tab 7 \tab 202 \tab - \tab 105 \tab - \tab 1473 \cr
   1985 \tab - \tab 59 \tab 1118 \tab 47 \tab 3 \tab 203 \tab - \tab 104 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 72 \tab 1368 \tab 50 \tab 19 \tab 208 \tab - \tab 102 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 44 \tab 720 \tab 36 \tab 4 \tab 120 \tab - \tab 53 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 40 \tab 727 \tab 49 \tab 4 \tab 154 \tab - \tab 59 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 42 \tab 665 \tab 44 \tab 1 \tab 112 \tab - \tab 53 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 38 \tab 699 \tab 60 \tab 7 \tab 148 \tab - \tab 41 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 44 \tab 671 \tab 56 \tab 7 \tab 223 \tab - \tab 74 \tab - \tab 1606 \cr
   1994 \tab 996 \tab 76 \tab 1253 \tab 107 \tab 5 \tab 439 \tab - \tab 116 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 93 \tab 1078 \tab 124 \tab 15 \tab 503 \tab - \tab 110 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 100 \tab 1017 \tab 121 \tab 8 \tab 515 \tab - \tab 121 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 76 \tab 997 \tab 153 \tab 11 \tab 489 \tab - \tab 135 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 43 \tab 486 \tab 35 \tab 5 \tab 292 \tab - \tab 63 \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 41 \tab 500 \tab 24 \tab 6 \tab 267 \tab - \tab 60 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 96 \tab 1052 \tab 78 \tab 17 \tab 624 \tab - \tab 136 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 43 \tab 654 \tab 68 \tab 15 \tab 484 \tab - \tab 88 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 52 \tab 555 \tab 41 \tab 17 \tab 519 \tab - \tab 97 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 37 \tab 567 \tab 50 \tab 19 \tab 553 \tab - \tab 82 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 55 \tab 669 \tab 44 \tab 19 \tab 812 \tab - \tab 115 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 65 \tab 700 \tab 50 \tab 34 \tab 940 \tab - \tab 101 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 55 \tab 499 \tab 56 \tab 23 \tab 846 \tab - \tab 95 \tab - \tab 2348 \cr
   2021 \tab 1358 \tab 112 \tab 693 \tab 9 \tab 8 \tab 1621 \tab - \tab 185 \tab 46 \tab 4032 \cr
   2022 \tab 1203 \tab 85 \tab 641 \tab 58 \tab 25 \tab 1377 \tab - \tab 146 \tab 9 \tab 3544 \cr
   Total \tab 27603 \tab 1953 \tab 24803 \tab 1810 \tab 321 \tab 12878 \tab 82 \tab 2885 \tab 55 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} always wrong
\item \code{2} almost always wrong
\item \code{3} wrong only sometimes
\item \code{4} not wrong at all
\item \code{5} other
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
