% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_01.R
\name{hrs1}
\alias{hrs1}
\title{Number of hours worked last week}
\source{
General Social Survey https://gss.norc.org
}
\description{
hrs1
}
\details{
Question 1a. If working, full or part time: how many hours did you work last week, at all jobs?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 6 \tab 60 \tab 61 \tab 65 \tab 66 \tab 7 \tab 70 \tab 71 \tab 72 \tab 75 \tab 8 \tab 80 \tab 84 \tab 89+ hours \tab 9 \tab don't know \tab no answer \tab 11 \tab 2 \tab 62 \tab 68 \tab 85 \tab 59 \tab 64 \tab 77 \tab 78 \tab 63 \tab 67 \tab 73 \tab 69 \tab 74 \tab 76 \tab 81 \tab 88 \tab 79 \tab 86 \tab 83 \tab 87 \tab 82 \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 719 \tab 1 \tab 1 \tab 4 \tab 10 \tab 2 \tab 2 \tab 10 \tab 10 \tab 1 \tab 5 \tab 1 \tab 23 \tab 3 \tab 3 \tab 4 \tab 11 \tab 10 \tab 2 \tab 5 \tab 8 \tab 1 \tab 1 \tab 21 \tab 2 \tab 9 \tab 1 \tab 3 \tab 36 \tab 8 \tab 12 \tab 13 \tab 3 \tab 2 \tab 303 \tab 3 \tab 13 \tab 7 \tab 10 \tab 24 \tab 4 \tab 1 \tab 31 \tab 3 \tab 1 \tab 32 \tab 1 \tab 12 \tab 2 \tab 2 \tab 17 \tab 4 \tab 1 \tab 2 \tab 1 \tab 32 \tab 1 \tab 5 \tab 3 \tab 2 \tab 14 \tab 1 \tab 4 \tab 2 \tab 5 \tab 5 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 741 \tab - \tab - \tab 5 \tab 4 \tab 2 \tab 2 \tab 15 \tab 10 \tab 2 \tab 4 \tab 1 \tab 20 \tab 2 \tab 2 \tab 1 \tab 10 \tab 9 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 22 \tab - \tab 11 \tab 1 \tab 2 \tab 33 \tab 5 \tab 12 \tab 1 \tab 3 \tab - \tab 329 \tab - \tab 13 \tab 6 \tab 12 \tab 26 \tab 7 \tab 2 \tab 27 \tab 2 \tab - \tab 44 \tab - \tab 4 \tab 3 \tab 3 \tab 10 \tab 7 \tab 1 \tab 2 \tab 2 \tab 20 \tab 1 \tab 4 \tab 1 \tab - \tab 6 \tab - \tab 2 \tab 3 \tab 4 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 723 \tab - \tab - \tab 5 \tab 4 \tab 1 \tab 1 \tab 13 \tab 20 \tab 2 \tab 1 \tab 1 \tab 19 \tab 7 \tab 2 \tab 6 \tab 11 \tab 10 \tab 6 \tab 1 \tab 5 \tab 1 \tab 2 \tab 27 \tab - \tab 21 \tab 6 \tab 2 \tab 28 \tab 1 \tab 12 \tab 8 \tab 3 \tab 3 \tab 305 \tab 3 \tab 11 \tab 6 \tab 9 \tab 30 \tab 7 \tab 5 \tab 23 \tab 2 \tab 1 \tab 42 \tab 1 \tab 6 \tab 2 \tab 3 \tab 10 \tab 5 \tab 1 \tab - \tab - \tab 26 \tab - \tab 2 \tab - \tab 1 \tab 6 \tab - \tab 2 \tab - \tab 7 \tab 4 \tab - \tab 6 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 749 \tab - \tab - \tab 7 \tab 3 \tab - \tab 1 \tab 21 \tab 11 \tab 1 \tab 5 \tab - \tab 15 \tab 1 \tab 3 \tab - \tab 10 \tab 16 \tab - \tab 2 \tab 7 \tab - \tab 1 \tab 23 \tab - \tab 23 \tab 1 \tab 2 \tab 32 \tab 6 \tab 12 \tab 9 \tab 3 \tab 1 \tab 310 \tab 4 \tab 7 \tab 7 \tab 5 \tab 24 \tab 3 \tab 3 \tab 32 \tab 2 \tab 2 \tab 30 \tab 2 \tab - \tab - \tab 3 \tab 14 \tab 10 \tab 2 \tab - \tab - \tab 29 \tab - \tab 5 \tab 2 \tab 2 \tab 8 \tab - \tab 1 \tab 1 \tab 4 \tab 3 \tab - \tab 6 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 648 \tab - \tab - \tab 11 \tab 5 \tab 1 \tab 2 \tab 10 \tab 6 \tab 2 \tab 1 \tab 1 \tab 18 \tab 4 \tab 2 \tab - \tab 10 \tab 11 \tab - \tab 1 \tab 4 \tab 2 \tab - \tab 25 \tab 2 \tab 16 \tab 3 \tab 1 \tab 33 \tab 10 \tab 7 \tab 21 \tab 3 \tab 2 \tab 379 \tab 5 \tab 12 \tab 5 \tab 14 \tab 33 \tab 7 \tab 3 \tab 37 \tab 5 \tab 4 \tab 45 \tab 2 \tab 4 \tab 1 \tab 5 \tab 15 \tab 9 \tab - \tab 2 \tab 1 \tab 34 \tab - \tab 6 \tab 1 \tab - \tab 7 \tab - \tab 2 \tab 2 \tab 3 \tab 8 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 677 \tab - \tab - \tab 6 \tab 6 \tab 4 \tab 1 \tab 10 \tab 9 \tab 1 \tab 3 \tab 1 \tab 25 \tab 3 \tab - \tab 2 \tab 8 \tab 7 \tab 2 \tab 2 \tab 5 \tab - \tab 3 \tab 29 \tab 2 \tab 11 \tab 1 \tab 1 \tab 33 \tab 8 \tab 15 \tab 23 \tab 2 \tab 1 \tab 330 \tab 7 \tab 18 \tab 4 \tab 7 \tab 29 \tab 9 \tab 5 \tab 38 \tab 1 \tab 2 \tab 42 \tab 3 \tab 7 \tab 2 \tab 5 \tab 15 \tab 7 \tab 1 \tab 1 \tab 3 \tab 29 \tab - \tab 6 \tab 1 \tab 2 \tab 9 \tab - \tab 7 \tab 5 \tab 4 \tab 4 \tab 6 \tab 6 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 644 \tab - \tab - \tab 9 \tab 5 \tab - \tab 2 \tab 11 \tab 6 \tab 2 \tab 4 \tab 1 \tab 25 \tab 3 \tab 1 \tab 1 \tab 10 \tab 17 \tab - \tab 5 \tab 3 \tab - \tab 2 \tab 19 \tab 3 \tab 12 \tab - \tab 5 \tab 29 \tab 8 \tab 12 \tab 8 \tab 2 \tab 1 \tab 318 \tab 6 \tab 16 \tab 6 \tab 12 \tab 41 \tab 8 \tab 3 \tab 24 \tab 2 \tab 1 \tab 46 \tab 2 \tab 2 \tab 4 \tab 2 \tab 10 \tab 5 \tab - \tab - \tab 3 \tab 36 \tab - \tab 12 \tab 1 \tab 1 \tab 10 \tab - \tab 3 \tab 3 \tab 5 \tab 11 \tab - \tab 11 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 836 \tab - \tab - \tab 7 \tab 5 \tab 4 \tab 2 \tab 14 \tab 15 \tab - \tab 5 \tab - \tab 30 \tab 4 \tab 1 \tab 4 \tab 13 \tab 16 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 39 \tab 1 \tab 25 \tab 5 \tab 3 \tab 42 \tab 9 \tab 18 \tab 28 \tab 7 \tab 3 \tab 391 \tab 3 \tab 19 \tab 6 \tab 10 \tab 44 \tab 6 \tab 4 \tab 30 \tab 1 \tab 4 \tab 43 \tab 4 \tab 8 \tab 3 \tab 1 \tab 18 \tab 7 \tab 1 \tab 3 \tab 5 \tab 43 \tab - \tab 11 \tab - \tab 1 \tab 6 \tab - \tab 2 \tab 1 \tab 8 \tab 8 \tab 1 \tab 6 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 688 \tab - \tab 1 \tab 4 \tab 15 \tab 1 \tab 1 \tab 10 \tab 6 \tab 2 \tab 2 \tab - \tab 28 \tab 3 \tab 1 \tab - \tab 13 \tab 11 \tab 4 \tab - \tab 3 \tab 2 \tab 3 \tab 41 \tab 2 \tab 27 \tab 2 \tab - \tab 32 \tab 10 \tab 19 \tab 23 \tab 6 \tab 4 \tab 339 \tab 2 \tab 13 \tab 4 \tab 8 \tab 37 \tab 5 \tab 3 \tab 33 \tab 1 \tab 2 \tab 47 \tab 3 \tab 1 \tab 4 \tab 4 \tab 11 \tab 7 \tab 1 \tab 3 \tab 5 \tab 37 \tab - \tab 10 \tab 2 \tab - \tab 10 \tab - \tab 5 \tab 2 \tab 4 \tab 6 \tab 3 \tab 7 \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 594 \tab - \tab 2 \tab 6 \tab 1 \tab - \tab 1 \tab 11 \tab 20 \tab 1 \tab 4 \tab - \tab 35 \tab 2 \tab 2 \tab 3 \tab 10 \tab 18 \tab - \tab 4 \tab 8 \tab 1 \tab 1 \tab 23 \tab 2 \tab 17 \tab - \tab - \tab 25 \tab 8 \tab 11 \tab 14 \tab 5 \tab 1 \tab 327 \tab 4 \tab 12 \tab 13 \tab 10 \tab 39 \tab 7 \tab 5 \tab 21 \tab 2 \tab 3 \tab 66 \tab 1 \tab 6 \tab 1 \tab 1 \tab 14 \tab 9 \tab 1 \tab 2 \tab 2 \tab 37 \tab - \tab 3 \tab - \tab 2 \tab 11 \tab - \tab 5 \tab 8 \tab 7 \tab 3 \tab 1 \tab 9 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1473 \cr
   1985 \tab 622 \tab 1 \tab - \tab 8 \tab 5 \tab - \tab 2 \tab 7 \tab 8 \tab 1 \tab 4 \tab 1 \tab 30 \tab 2 \tab 4 \tab 2 \tab 8 \tab 22 \tab 1 \tab 3 \tab 3 \tab 2 \tab 1 \tab 18 \tab 3 \tab 14 \tab 3 \tab 6 \tab 17 \tab 16 \tab 12 \tab 26 \tab 4 \tab 5 \tab 324 \tab 4 \tab 16 \tab 4 \tab 13 \tab 32 \tab 16 \tab 6 \tab 27 \tab 3 \tab 4 \tab 53 \tab 5 \tab 9 \tab 3 \tab 4 \tab 16 \tab 2 \tab 1 \tab 2 \tab 5 \tab 52 \tab - \tab 9 \tab - \tab 1 \tab 12 \tab - \tab 3 \tab 5 \tab 9 \tab 10 \tab 7 \tab 7 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 626 \tab - \tab - \tab 5 \tab 5 \tab 1 \tab 2 \tab 10 \tab 7 \tab - \tab 2 \tab - \tab 21 \tab 2 \tab 1 \tab 1 \tab 8 \tab 13 \tab - \tab 2 \tab 5 \tab 2 \tab 1 \tab 29 \tab 2 \tab 16 \tab 1 \tab 2 \tab 32 \tab 7 \tab 8 \tab 16 \tab 2 \tab 1 \tab 294 \tab 6 \tab 7 \tab 5 \tab 10 \tab 39 \tab 8 \tab 3 \tab 21 \tab 1 \tab 2 \tab 67 \tab 5 \tab 8 \tab 5 \tab 1 \tab 21 \tab 8 \tab 1 \tab 1 \tab 2 \tab 45 \tab 1 \tab 8 \tab 1 \tab 4 \tab 14 \tab - \tab 3 \tab 4 \tab 8 \tab 6 \tab 1 \tab 7 \tab 1 \tab 5 \tab 4 \tab - \tab 1 \tab 7 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1470 \cr
   1987 \tab 684 \tab - \tab - \tab 11 \tab 10 \tab 1 \tab 2 \tab 11 \tab 10 \tab 5 \tab 5 \tab - \tab 36 \tab 2 \tab 5 \tab 2 \tab 19 \tab 22 \tab 2 \tab 1 \tab 12 \tab 2 \tab - \tab 29 \tab - \tab 26 \tab 3 \tab 2 \tab 38 \tab 15 \tab 11 \tab 23 \tab 3 \tab 1 \tab 392 \tab 3 \tab 10 \tab 10 \tab 14 \tab 52 \tab 13 \tab 8 \tab 43 \tab 9 \tab 3 \tab 65 \tab 2 \tab 8 \tab 2 \tab 4 \tab 22 \tab 10 \tab - \tab 5 \tab 1 \tab 59 \tab 1 \tab 12 \tab 1 \tab 5 \tab 10 \tab - \tab 3 \tab 6 \tab 8 \tab 9 \tab 2 \tab 10 \tab 2 \tab - \tab 7 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 587 \tab 3 \tab - \tab 5 \tab 6 \tab 3 \tab 2 \tab 13 \tab 7 \tab 2 \tab 1 \tab 1 \tab 26 \tab 1 \tab 2 \tab 1 \tab 11 \tab 12 \tab 1 \tab 2 \tab 6 \tab 1 \tab - \tab 26 \tab 1 \tab 16 \tab 3 \tab - \tab 21 \tab 13 \tab 10 \tab 18 \tab 1 \tab - \tab 330 \tab 1 \tab 11 \tab 6 \tab 5 \tab 49 \tab 8 \tab 4 \tab 26 \tab 2 \tab 3 \tab 65 \tab 2 \tab 4 \tab 1 \tab 4 \tab 26 \tab 13 \tab - \tab - \tab 4 \tab 47 \tab - \tab 11 \tab 3 \tab 1 \tab 15 \tab - \tab 1 \tab 2 \tab 7 \tab 7 \tab 1 \tab 4 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 624 \tab - \tab - \tab 10 \tab 4 \tab 1 \tab 1 \tab 7 \tab 13 \tab 1 \tab 1 \tab 1 \tab 27 \tab 4 \tab 2 \tab 5 \tab 9 \tab 21 \tab - \tab 2 \tab 9 \tab 1 \tab 2 \tab 30 \tab 2 \tab 14 \tab 1 \tab 2 \tab 23 \tab 6 \tab 9 \tab 21 \tab 2 \tab 2 \tab 304 \tab 4 \tab 15 \tab 9 \tab 7 \tab 47 \tab 6 \tab 4 \tab 23 \tab 5 \tab 1 \tab 69 \tab 3 \tab 8 \tab 3 \tab 5 \tab 24 \tab 8 \tab 2 \tab 3 \tab 5 \tab 51 \tab 2 \tab 6 \tab - \tab 1 \tab 10 \tab - \tab 4 \tab 5 \tab 7 \tab 8 \tab 1 \tab 6 \tab 3 \tab 2 \tab 4 \tab 2 \tab - \tab 5 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1537 \cr
   1990 \tab 521 \tab 1 \tab 1 \tab 10 \tab 7 \tab - \tab 2 \tab 8 \tab 5 \tab 1 \tab 2 \tab 1 \tab 28 \tab 2 \tab 4 \tab - \tab 11 \tab 8 \tab 4 \tab 1 \tab 8 \tab 1 \tab 2 \tab 31 \tab 1 \tab 15 \tab 3 \tab 1 \tab 27 \tab 6 \tab 13 \tab 26 \tab 1 \tab 2 \tab 279 \tab 2 \tab 12 \tab 8 \tab 10 \tab 36 \tab 15 \tab 8 \tab 25 \tab 4 \tab 6 \tab 62 \tab 1 \tab 7 \tab 2 \tab 3 \tab 17 \tab 7 \tab 2 \tab 2 \tab 2 \tab 36 \tab 1 \tab 17 \tab - \tab 1 \tab 14 \tab - \tab 1 \tab 3 \tab 6 \tab 2 \tab 4 \tab 2 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1372 \cr
   1991 \tab 633 \tab 1 \tab - \tab 10 \tab 7 \tab 1 \tab 1 \tab 16 \tab 10 \tab 1 \tab 6 \tab 2 \tab 35 \tab 4 \tab 2 \tab 3 \tab 9 \tab 11 \tab 3 \tab - \tab 5 \tab 3 \tab 1 \tab 31 \tab 3 \tab 24 \tab 3 \tab 4 \tab 22 \tab 8 \tab 8 \tab 11 \tab 3 \tab 4 \tab 300 \tab 4 \tab 9 \tab 4 \tab 7 \tab 38 \tab 8 \tab 4 \tab 17 \tab 6 \tab 4 \tab 70 \tab 3 \tab 9 \tab 2 \tab 3 \tab 17 \tab 6 \tab 1 \tab 1 \tab 5 \tab 44 \tab 1 \tab 10 \tab - \tab - \tab 7 \tab 1 \tab 4 \tab 3 \tab 7 \tab 9 \tab 2 \tab 7 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 4 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 630 \tab - \tab - \tab 7 \tab 5 \tab 2 \tab 3 \tab 11 \tab 12 \tab 1 \tab 4 \tab - \tab 35 \tab 7 \tab 3 \tab 3 \tab 14 \tab 15 \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 25 \tab - \tab 14 \tab 2 \tab 3 \tab 29 \tab 11 \tab 9 \tab 16 \tab 2 \tab 2 \tab 296 \tab 4 \tab 16 \tab 6 \tab 18 \tab 48 \tab 11 \tab 3 \tab 21 \tab 1 \tab 3 \tab 89 \tab 2 \tab 10 \tab 2 \tab 3 \tab 40 \tab 9 \tab 2 \tab 4 \tab 2 \tab 55 \tab 1 \tab 18 \tab 1 \tab 3 \tab 17 \tab - \tab 1 \tab 6 \tab 6 \tab 16 \tab - \tab 2 \tab 1 \tab 2 \tab 6 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1081 \tab - \tab - \tab 17 \tab 13 \tab 6 \tab 3 \tab 10 \tab 18 \tab - \tab 4 \tab 3 \tab 56 \tab 1 \tab 10 \tab 6 \tab 32 \tab 33 \tab 5 \tab 2 \tab 10 \tab 2 \tab 1 \tab 59 \tab 5 \tab 33 \tab 6 \tab 5 \tab 42 \tab 25 \tab 19 \tab 42 \tab 3 \tab 7 \tab 645 \tab 8 \tab 17 \tab 19 \tab 21 \tab 104 \tab 18 \tab 12 \tab 59 \tab 5 \tab 5 \tab 123 \tab 7 \tab 17 \tab 6 \tab 8 \tab 41 \tab 17 \tab 1 \tab 9 \tab 10 \tab 126 \tab - \tab 23 \tab 2 \tab 3 \tab 23 \tab - \tab 6 \tab 4 \tab 12 \tab 20 \tab 2 \tab 14 \tab 3 \tab 2 \tab 9 \tab 2 \tab 2 \tab 5 \tab 2 \tab 1 \tab 2 \tab 5 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2992 \cr
   1996 \tab 961 \tab - \tab - \tab 20 \tab 10 \tab 1 \tab 2 \tab 17 \tab 10 \tab 3 \tab 8 \tab 5 \tab 61 \tab 4 \tab 3 \tab 4 \tab 15 \tab 29 \tab 4 \tab 7 \tab 7 \tab 1 \tab 1 \tab 61 \tab 3 \tab 41 \tab 5 \tab 4 \tab 57 \tab 29 \tab 17 \tab 29 \tab 10 \tab 7 \tab 640 \tab 10 \tab 38 \tab 14 \tab 28 \tab 107 \tab 17 \tab 5 \tab 41 \tab 8 \tab 4 \tab 159 \tab 2 \tab 18 \tab 5 \tab 9 \tab 58 \tab 11 \tab 4 \tab 8 \tab 3 \tab 105 \tab - \tab 35 \tab 2 \tab 2 \tab 36 \tab - \tab 3 \tab 6 \tab 6 \tab 18 \tab 1 \tab 16 \tab 3 \tab 1 \tab 7 \tab 3 \tab 3 \tab 2 \tab 3 \tab 1 \tab 4 \tab 5 \tab 2 \tab - \tab 3 \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 2904 \cr
   1998 \tab 959 \tab - \tab 2 \tab 12 \tab 8 \tab 2 \tab 4 \tab 24 \tab 16 \tab 2 \tab 8 \tab 3 \tab 50 \tab 4 \tab 5 \tab 6 \tab 19 \tab 28 \tab 6 \tab 2 \tab 13 \tab 3 \tab 2 \tab 63 \tab 2 \tab 37 \tab 2 \tab 8 \tab 51 \tab 21 \tab 14 \tab 33 \tab 6 \tab 9 \tab 605 \tab 12 \tab 30 \tab 18 \tab 18 \tab 96 \tab 18 \tab 12 \tab 56 \tab 6 \tab 3 \tab 175 \tab 2 \tab 19 \tab 6 \tab 8 \tab 34 \tab 23 \tab 2 \tab 5 \tab 7 \tab 102 \tab - \tab 22 \tab 1 \tab 2 \tab 35 \tab 1 \tab 4 \tab 6 \tab 11 \tab 14 \tab 2 \tab 11 \tab 1 \tab - \tab 17 \tab 2 \tab - \tab 3 \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 987 \tab - \tab - \tab 16 \tab 7 \tab 4 \tab 2 \tab 26 \tab 9 \tab 3 \tab 4 \tab 2 \tab 56 \tab 5 \tab 6 \tab 4 \tab 9 \tab 29 \tab 5 \tab 8 \tab 11 \tab 1 \tab 3 \tab 53 \tab 1 \tab 30 \tab 2 \tab 2 \tab 50 \tab 19 \tab 22 \tab 29 \tab 7 \tab 4 \tab 644 \tab 7 \tab 27 \tab 20 \tab 30 \tab 118 \tab 16 \tab 8 \tab 46 \tab 8 \tab 3 \tab 156 \tab 3 \tab 16 \tab 3 \tab 7 \tab 43 \tab 11 \tab 2 \tab 4 \tab 5 \tab 97 \tab - \tab 21 \tab 1 \tab 1 \tab 24 \tab - \tab 7 \tab 3 \tab 8 \tab 15 \tab 1 \tab 11 \tab - \tab 1 \tab 11 \tab - \tab - \tab 5 \tab 2 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2817 \cr
   2002 \tab 1021 \tab - \tab 1 \tab 12 \tab 15 \tab 1 \tab 5 \tab 11 \tab 24 \tab 3 \tab 6 \tab - \tab 63 \tab 1 \tab 4 \tab 4 \tab 21 \tab 26 \tab 5 \tab 5 \tab 9 \tab 2 \tab 4 \tab 66 \tab 6 \tab 36 \tab 2 \tab 7 \tab 46 \tab 21 \tab 24 \tab 27 \tab 6 \tab 5 \tab 532 \tab 8 \tab 31 \tab 21 \tab 21 \tab 110 \tab 8 \tab 10 \tab 44 \tab 7 \tab 4 \tab 137 \tab 3 \tab 15 \tab 7 \tab 5 \tab 47 \tab 10 \tab - \tab 3 \tab 5 \tab 94 \tab 2 \tab 18 \tab 4 \tab 2 \tab 30 \tab - \tab 7 \tab 8 \tab 4 \tab 9 \tab 6 \tab 17 \tab 3 \tab 3 \tab 12 \tab 1 \tab 3 \tab 6 \tab 3 \tab 3 \tab - \tab 4 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab 1025 \tab - \tab 3 \tab 14 \tab 15 \tab 2 \tab 4 \tab 14 \tab 18 \tab 2 \tab 6 \tab 2 \tab 53 \tab 2 \tab 4 \tab 3 \tab 29 \tab 32 \tab 5 \tab 2 \tab 10 \tab 1 \tab 2 \tab 61 \tab 3 \tab 33 \tab 2 \tab 6 \tab 51 \tab 22 \tab 15 \tab 34 \tab 7 \tab 5 \tab 554 \tab 4 \tab 28 \tab 15 \tab 23 \tab 88 \tab 21 \tab 12 \tab 37 \tab 4 \tab 3 \tab 148 \tab 1 \tab 17 \tab 7 \tab 1 \tab 43 \tab 15 \tab 4 \tab 6 \tab 4 \tab 106 \tab 2 \tab 25 \tab 1 \tab - \tab 27 \tab - \tab 5 \tab 8 \tab 12 \tab 30 \tab 4 \tab 14 \tab 1 \tab 4 \tab 20 \tab 1 \tab 4 \tab 3 \tab 3 \tab 4 \tab 1 \tab 5 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 1745 \tab - \tab 4 \tab 19 \tab 13 \tab 7 \tab 11 \tab 22 \tab 20 \tab 4 \tab 9 \tab 4 \tab 81 \tab 6 \tab 6 \tab 3 \tab 37 \tab 41 \tab 6 \tab 10 \tab 12 \tab 6 \tab 2 \tab 85 \tab 4 \tab 52 \tab 7 \tab 9 \tab 88 \tab 24 \tab 25 \tab 52 \tab 8 \tab 8 \tab 888 \tab 8 \tab 54 \tab 32 \tab 27 \tab 147 \tab 23 \tab 17 \tab 64 \tab 7 \tab 5 \tab 243 \tab 7 \tab 18 \tab 14 \tab 10 \tab 81 \tab 13 \tab 4 \tab 12 \tab 9 \tab 163 \tab - \tab 30 \tab 3 \tab 3 \tab 43 \tab - \tab 7 \tab 17 \tab 15 \tab 31 \tab 1 \tab 20 \tab 4 \tab 9 \tab 17 \tab 4 \tab 1 \tab 7 \tab 1 \tab 3 \tab - \tab 4 \tab - \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 4510 \cr
   2008 \tab 807 \tab - \tab 4 \tab 5 \tab 10 \tab - \tab 2 \tab 9 \tab 6 \tab 4 \tab 3 \tab 3 \tab 34 \tab 4 \tab 6 \tab 4 \tab 21 \tab 23 \tab 4 \tab 6 \tab 3 \tab 1 \tab 1 \tab 47 \tab 1 \tab 23 \tab 6 \tab 3 \tab 32 \tab 13 \tab 14 \tab 17 \tab 5 \tab 5 \tab 371 \tab 6 \tab 24 \tab 15 \tab 13 \tab 65 \tab 17 \tab 9 \tab 30 \tab 2 \tab 1 \tab 104 \tab 1 \tab 12 \tab - \tab 2 \tab 18 \tab 13 \tab - \tab 4 \tab 4 \tab 68 \tab 1 \tab 20 \tab 1 \tab 1 \tab 14 \tab - \tab 2 \tab 3 \tab 6 \tab 20 \tab - \tab 9 \tab 1 \tab 3 \tab 10 \tab - \tab 1 \tab 5 \tab 2 \tab 2 \tab 2 \tab 8 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 890 \tab - \tab 1 \tab 9 \tab 6 \tab 2 \tab 3 \tab 10 \tab 11 \tab - \tab 3 \tab 1 \tab 42 \tab 2 \tab 6 \tab - \tab 15 \tab 22 \tab 4 \tab 8 \tab 8 \tab 2 \tab 2 \tab 46 \tab 1 \tab 26 \tab 2 \tab 3 \tab 34 \tab 13 \tab 16 \tab 23 \tab 4 \tab 3 \tab 335 \tab 7 \tab 24 \tab 7 \tab 10 \tab 56 \tab 8 \tab 5 \tab 22 \tab 3 \tab 2 \tab 118 \tab 2 \tab 24 \tab - \tab 3 \tab 27 \tab 9 \tab 1 \tab 2 \tab 3 \tab 59 \tab - \tab 10 \tab - \tab 2 \tab 19 \tab - \tab 4 \tab 2 \tab 7 \tab 6 \tab 1 \tab 12 \tab 2 \tab 3 \tab 12 \tab 1 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2044 \cr
   2012 \tab 835 \tab - \tab 2 \tab 6 \tab 7 \tab - \tab 1 \tab 13 \tab 14 \tab 2 \tab 5 \tab 3 \tab 46 \tab 3 \tab 3 \tab 1 \tab 15 \tab 17 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 47 \tab 2 \tab 25 \tab 3 \tab 6 \tab 32 \tab 17 \tab 11 \tab 21 \tab 4 \tab 6 \tab 349 \tab 6 \tab 18 \tab 10 \tab 14 \tab 57 \tab 14 \tab 6 \tab 28 \tab 3 \tab 3 \tab 76 \tab - \tab 7 \tab 5 \tab - \tab 33 \tab 4 \tab 3 \tab 4 \tab - \tab 75 \tab - \tab 13 \tab - \tab 1 \tab 21 \tab - \tab 3 \tab 2 \tab 11 \tab 15 \tab - \tab 8 \tab 1 \tab - \tab 13 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1033 \tab - \tab 5 \tab 10 \tab 8 \tab - \tab 2 \tab 21 \tab 8 \tab 2 \tab 4 \tab - \tab 52 \tab 5 \tab 2 \tab 2 \tab 19 \tab 24 \tab 3 \tab 5 \tab 14 \tab 2 \tab 3 \tab 65 \tab 2 \tab 24 \tab 3 \tab 7 \tab 42 \tab 18 \tab 18 \tab 20 \tab 9 \tab 3 \tab 507 \tab 4 \tab 19 \tab 6 \tab 12 \tab 70 \tab 8 \tab 12 \tab 25 \tab 4 \tab 2 \tab 140 \tab 3 \tab 12 \tab 3 \tab 4 \tab 34 \tab 8 \tab 8 \tab 4 \tab 1 \tab 102 \tab - \tab 19 \tab 1 \tab 1 \tab 27 \tab - \tab 2 \tab 5 \tab 11 \tab 15 \tab 1 \tab 15 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2538 \cr
   2016 \tab 1198 \tab - \tab 5 \tab 16 \tab 8 \tab - \tab 2 \tab 16 \tab 18 \tab 4 \tab 6 \tab 3 \tab 46 \tab 4 \tab 8 \tab 5 \tab 11 \tab 32 \tab 9 \tab 9 \tab 13 \tab 2 \tab 3 \tab 62 \tab 2 \tab 39 \tab 9 \tab 9 \tab 45 \tab 27 \tab 14 \tab 17 \tab 10 \tab 11 \tab 526 \tab 12 \tab 30 \tab 12 \tab 11 \tab 92 \tab 13 \tab 10 \tab 38 \tab 4 \tab 5 \tab 138 \tab 3 \tab 11 \tab 6 \tab 8 \tab 37 \tab 12 \tab 3 \tab 2 \tab 5 \tab 82 \tab 1 \tab 16 \tab - \tab 3 \tab 30 \tab - \tab 5 \tab 6 \tab 10 \tab 9 \tab 4 \tab 10 \tab 1 \tab 4 \tab 19 \tab 1 \tab 1 \tab 4 \tab 5 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 953 \tab - \tab 7 \tab 9 \tab 4 \tab 3 \tab 1 \tab 17 \tab 17 \tab 1 \tab 3 \tab 1 \tab 47 \tab 2 \tab 8 \tab 1 \tab 14 \tab 28 \tab 2 \tab 4 \tab 9 \tab 3 \tab 5 \tab 62 \tab - \tab 29 \tab 4 \tab 5 \tab 49 \tab 26 \tab 17 \tab 19 \tab 1 \tab - \tab 420 \tab 5 \tab 21 \tab 11 \tab 17 \tab 80 \tab 13 \tab 4 \tab 28 \tab 2 \tab 2 \tab 112 \tab 2 \tab 9 \tab 6 \tab 10 \tab 35 \tab 3 \tab 5 \tab 2 \tab 8 \tab 90 \tab 1 \tab 17 \tab - \tab 2 \tab 25 \tab - \tab 5 \tab 5 \tab 6 \tab 4 \tab 2 \tab 12 \tab - \tab 3 \tab 11 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1883 \tab 10 \tab 3 \tab 31 \tab 10 \tab 2 \tab 3 \tab 27 \tab 16 \tab 3 \tab 3 \tab 1 \tab 81 \tab 3 \tab 3 \tab 2 \tab 25 \tab 35 \tab 2 \tab 4 \tab 8 \tab 3 \tab 3 \tab 59 \tab - \tab 26 \tab 8 \tab 4 \tab 48 \tab 31 \tab 18 \tab 29 \tab 4 \tab 5 \tab 878 \tab 10 \tab 30 \tab 20 \tab 15 \tab 146 \tab 12 \tab 4 \tab 34 \tab 3 \tab 5 \tab 173 \tab - \tab 16 \tab 2 \tab 3 \tab 50 \tab 11 \tab 4 \tab 4 \tab 3 \tab 81 \tab - \tab 8 \tab 1 \tab 2 \tab 21 \tab - \tab 5 \tab 2 \tab 14 \tab 16 \tab 3 \tab 8 \tab 4 \tab 2 \tab 25 \tab 1 \tab 3 \tab 3 \tab 2 \tab - \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 6 \tab 4032 \cr
   2022 \tab 1593 \tab 19 \tab 3 \tab 20 \tab 14 \tab - \tab 5 \tab 23 \tab 13 \tab 2 \tab 12 \tab - \tab 70 \tab 2 \tab 4 \tab 2 \tab 19 \tab 33 \tab 5 \tab 3 \tab 12 \tab 2 \tab 1 \tab 64 \tab 4 \tab 30 \tab 4 \tab 11 \tab 45 \tab 26 \tab 13 \tab 23 \tab 4 \tab 6 \tab 775 \tab 12 \tab 21 \tab 9 \tab 16 \tab 123 \tab 12 \tab 5 \tab 29 \tab 3 \tab 5 \tab 148 \tab 2 \tab 12 \tab 4 \tab 9 \tab 44 \tab 9 \tab 4 \tab 7 \tab 6 \tab 82 \tab - \tab 16 \tab - \tab 3 \tab 16 \tab - \tab 4 \tab 5 \tab 8 \tab 18 \tab 3 \tab 13 \tab 2 \tab 3 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 7 \tab 3544 \cr
   Total \tab 30530 \tab 36 \tab 45 \tab 346 \tab 255 \tab 54 \tab 80 \tab 468 \tab 403 \tab 61 \tab 143 \tab 43 \tab 1314 \tab 104 \tab 118 \tab 85 \tab 496 \tab 681 \tab 96 \tab 111 \tab 244 \tab 55 \tab 55 \tab 1388 \tab 62 \tab 795 \tab 104 \tab 128 \tab 1244 \tab 487 \tab 467 \tab 720 \tab 143 \tab 119 \tab 14519 \tab 184 \tab 642 \tab 345 \tab 457 \tab 2127 \tab 366 \tab 205 \tab 1080 \tab 121 \tab 98 \tab 3127 \tab 80 \tab 336 \tab 116 \tab 143 \tab 942 \tab 302 \tab 63 \tab 109 \tab 121 \tab 2144 \tab 16 \tab 448 \tab 34 \tab 55 \tab 577 \tab 3 \tab 122 \tab 143 \tab 250 \tab 364 \tab 64 \tab 294 \tab 58 \tab 55 \tab 232 \tab 32 \tab 37 \tab 86 \tab 59 \tab 37 \tab 26 \tab 77 \tab 10 \tab 18 \tab 41 \tab 25 \tab 15 \tab 8 \tab 16 \tab 9 \tab 5 \tab 9 \tab 3 \tab 17 \tab 8 \tab 3 \tab 14 \tab 13 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{89} 89+ hours
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
