% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_371.R
\name{hsmath}
\alias{hsmath}
\title{hsmath}
\source{
General Social Survey https://gss.norc.org
}
\description{
The highest level of math r completed in high school
}
\details{
Question 1050. Now, let me ask you to think about the courses you took in high school. What was the highest level of math that you completed in high school?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab calculus \tab don't know \tab general math, business, or vocational math \tab geometry (plane or solid or both) \tab no answer \tab no math in hs; didn't go to hs \tab one year of algebra \tab other \tab pre-algebra \tab pre-calculus \tab statistics/probability \tab trigonometry/linear programming/analysis \tab two years of algebra \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2690 \tab 180 \tab 30 \tab 266 \tab 334 \tab 3 \tab 29 \tab 256 \tab 15 \tab 82 \tab 98 \tab 33 \tab 209 \tab 285 \tab 4510 \cr
   2008 \tab 597 \tab 115 \tab 38 \tab 230 \tab 236 \tab 2 \tab 24 \tab 201 \tab 11 \tab 57 \tab 92 \tab 25 \tab 183 \tab 212 \tab 2023 \cr
   2010 \tab 1370 \tab 82 \tab 20 \tab 60 \tab 96 \tab 64 \tab 3 \tab 76 \tab 10 \tab 17 \tab 40 \tab 17 \tab 80 \tab 109 \tab 2044 \cr
   2012 \tab 1068 \tab 106 \tab 42 \tab 114 \tab 142 \tab 3 \tab 8 \tab 126 \tab 3 \tab 23 \tab 65 \tab 25 \tab 91 \tab 158 \tab 1974 \cr
   2014 \tab 1348 \tab 122 \tab 43 \tab 159 \tab 187 \tab 4 \tab 25 \tab 168 \tab 1 \tab 50 \tab 65 \tab 16 \tab 121 \tab 229 \tab 2538 \cr
   2016 \tab 1532 \tab 137 \tab 50 \tab 173 \tab 227 \tab 10 \tab 19 \tab 185 \tab 7 \tab 58 \tab 86 \tab 28 \tab 128 \tab 227 \tab 2867 \cr
   2018 \tab 1222 \tab 118 \tab 60 \tab 117 \tab 163 \tab 8 \tab 19 \tab 176 \tab 6 \tab 33 \tab 85 \tab 29 \tab 121 \tab 191 \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 63913 \tab 860 \tab 283 \tab 1119 \tab 1385 \tab 94 \tab 127 \tab 1188 \tab 53 \tab 320 \tab 531 \tab 173 \tab 933 \tab 1411 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} no math in hs; didn't go to hs
\item \code{1} general math, business, or vocational math
\item \code{2} pre-algebra
\item \code{3} one year of algebra
\item \code{4} two years of algebra
\item \code{5} geometry (plane or solid or both)
\item \code{6} trigonometry/linear programming/analysis
\item \code{7} pre-calculus
\item \code{8} calculus
\item \code{9} statistics/probability
\item \code{10} other
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
