% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_77.R
\name{hunt}
\alias{hunt}
\title{Does r or spouse hunt}
\source{
General Social Survey https://gss.norc.org
}
\description{
hunt
}
\details{
Question 240a. Do you (or does you (husband/wife)) go hunting?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab no answer \tab no, neither r nor spouse hunts \tab yes, both do \tab yes, respondent does \tab yes, spouse does \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 3 \tab 1081 \tab 49 \tab 261 \tab 136 \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 3 \tab 1078 \tab 55 \tab 207 \tab 125 \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 10 \tab 1429 \tab 39 \tab 260 \tab 122 \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 6 \tab 1097 \tab 38 \tab 207 \tab 125 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 3 \tab 1158 \tab 46 \tab 235 \tab 92 \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 5 \tab 1426 \tab 55 \tab 223 \tab 110 \tab - \tab - \tab 1819 \cr
   1988 \tab 504 \tab 1 \tab 777 \tab 19 \tab 119 \tab 61 \tab - \tab - \tab 1481 \cr
   1989 \tab 504 \tab 3 \tab 811 \tab 34 \tab 124 \tab 61 \tab - \tab - \tab 1537 \cr
   1990 \tab 455 \tab 4 \tab 710 \tab 23 \tab 127 \tab 53 \tab - \tab - \tab 1372 \cr
   1991 \tab 524 \tab 6 \tab 781 \tab 21 \tab 131 \tab 54 \tab - \tab - \tab 1517 \cr
   1993 \tab 531 \tab 1 \tab 864 \tab 17 \tab 139 \tab 54 \tab - \tab - \tab 1606 \cr
   1994 \tab 996 \tab 4 \tab 1589 \tab 41 \tab 258 \tab 104 \tab - \tab - \tab 2992 \cr
   1996 \tab 981 \tab 2 \tab 1516 \tab 44 \tab 263 \tab 98 \tab - \tab - \tab 2904 \cr
   1998 \tab 950 \tab 5 \tab 1540 \tab 28 \tab 215 \tab 94 \tab - \tab - \tab 2832 \cr
   2000 \tab 956 \tab 5 \tab 1536 \tab 25 \tab 216 \tab 79 \tab - \tab - \tab 2817 \cr
   2002 \tab 1841 \tab - \tab 752 \tab 18 \tab 129 \tab 24 \tab 1 \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 1 \tab 729 \tab 24 \tab 106 \tab 38 \tab - \tab - \tab 2812 \cr
   2006 \tab 2507 \tab - \tab 1682 \tab 39 \tab 193 \tab 88 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 1 \tab 1124 \tab 21 \tab 149 \tab 57 \tab - \tab - \tab 2023 \cr
   2010 \tab 763 \tab 1 \tab 1052 \tab 17 \tab 167 \tab 44 \tab - \tab - \tab 2044 \cr
   2012 \tab 666 \tab - \tab 1062 \tab 30 \tab 144 \tab 72 \tab - \tab - \tab 1974 \cr
   2014 \tab 824 \tab 1 \tab 1450 \tab 40 \tab 157 \tab 66 \tab - \tab - \tab 2538 \cr
   2016 \tab 977 \tab 3 \tab 1576 \tab 46 \tab 195 \tab 69 \tab 1 \tab - \tab 2867 \cr
   2018 \tab 774 \tab - \tab 1312 \tab 30 \tab 170 \tab 61 \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab 3479 \tab 80 \tab 338 \tab 128 \tab - \tab 7 \tab 4032 \cr
   2022 \tab 1203 \tab 1 \tab 1996 \tab 41 \tab 242 \tab 57 \tab 3 \tab 1 \tab 3544 \cr
   Total \tab 30732 \tab 69 \tab 33607 \tab 920 \tab 4975 \tab 2072 \tab 7 \tab 8 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes, respondent does
\item \code{2} yes, spouse does
\item \code{3} yes, both do
\item \code{4} no, neither r nor spouse hunts
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
