% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_567.R
\name{incdef}
\alias{incdef}
\title{Distance below poverty line}
\source{
General Social Survey https://gss.norc.org
}
\description{
incdef
}
\details{
Question 1576b. Income Deficit of Household
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab -10,000 to -5,000 \tab -4,999 to -1,000 \tab -999 to 999 \tab 1,000 to 4,999 \tab 10,000 to 20,000 \tab 20,001 + \tab 5,000 to 9,999 \tab not imputable \tab < -10,000 \tab not available in this year \tab iap \tab Total \cr
   1972 \tab 6 \tab 139 \tab 168 \tab 388 \tab 304 \tab 63 \tab 400 \tab 145 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 4 \tab 89 \tab 156 \tab 325 \tab 298 \tab 94 \tab 418 \tab 120 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 5 \tab 91 \tab 147 \tab 265 \tab 326 \tab 116 \tab 401 \tab 133 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 3 \tab 100 \tab 159 \tab 289 \tab 307 \tab 128 \tab 417 \tab 87 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 4 \tab 93 \tab 196 \tab 259 \tab 323 \tab 148 \tab 366 \tab 110 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 7 \tab 65 \tab 147 \tab 258 \tab 370 \tab 212 \tab 330 \tab 141 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 3 \tab 101 \tab 154 \tab 269 \tab 405 \tab 221 \tab 272 \tab 106 \tab 1 \tab - \tab - \tab 1532 \cr
   1980 \tab 15 \tab 105 \tab 106 \tab 190 \tab 361 \tab 330 \tab 247 \tab 112 \tab 2 \tab - \tab - \tab 1468 \cr
   1982 \tab 68 \tab 156 \tab 152 \tab 239 \tab 339 \tab 432 \tab 277 \tab 189 \tab 8 \tab - \tab - \tab 1860 \cr
   1983 \tab 34 \tab 123 \tab 92 \tab 216 \tab 290 \tab 482 \tab 204 \tab 151 \tab 7 \tab - \tab - \tab 1599 \cr
   1984 \tab 36 \tab 132 \tab 81 \tab 142 \tab 315 \tab 418 \tab 211 \tab 134 \tab 4 \tab - \tab - \tab 1473 \cr
   1985 \tab 37 \tab 109 \tab 100 \tab 158 \tab 305 \tab 518 \tab 175 \tab 124 \tab 8 \tab - \tab - \tab 1534 \cr
   1986 \tab 42 \tab 112 \tab 87 \tab 132 \tab 289 \tab 499 \tab 163 \tab 134 \tab 12 \tab - \tab - \tab 1470 \cr
   1987 \tab 71 \tab 159 \tab 81 \tab 173 \tab 330 \tab 647 \tab 181 \tab 162 \tab 15 \tab - \tab - \tab 1819 \cr
   1988 \tab 52 \tab 93 \tab 80 \tab 141 \tab 274 \tab 557 \tab 152 \tab 125 \tab 7 \tab - \tab - \tab 1481 \cr
   1989 \tab 45 \tab 76 \tab 55 \tab 139 \tab 272 \tab 605 \tab 170 \tab 162 \tab 13 \tab - \tab - \tab 1537 \cr
   1990 \tab 32 \tab 81 \tab 63 \tab 105 \tab 258 \tab 552 \tab 130 \tab 146 \tab 5 \tab - \tab - \tab 1372 \cr
   1991 \tab 92 \tab 118 \tab 63 \tab 133 \tab 175 \tab 588 \tab 148 \tab 158 \tab 42 \tab - \tab - \tab 1517 \cr
   1993 \tab 106 \tab 107 \tab 55 \tab 132 \tab 163 \tab 741 \tab 107 \tab 152 \tab 43 \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 662 \tab 2049 \tab 2142 \tab 3953 \tab 5704 \tab 7351 \tab 4769 \tab 2591 \tab 167 \tab 35426 \tab 7576 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} < -10,000
\item \code{2} -10,000 to -5,000
\item \code{3} -4,999 to -1,000
\item \code{4} -999 to 999
\item \code{5} 1,000 to 4,999
\item \code{6} 5,000 to 9,999
\item \code{7} 10,000 to 20,000
\item \code{8} 20,001 +
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
