% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_14.R
\name{incom16}
\alias{incom16}
\title{R's family income when 16 yrs old}
\source{
General Social Survey https://gss.norc.org
}
\description{
incom16
}
\details{
Question 30. Thinking about the time when you were 16 years old, compared with American families in general then, would you say your family income was far below average, below average, average, above average, or far above average?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab above average \tab average \tab below average \tab don't know \tab far above average \tab far below average \tab lived in institution \tab no answer \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 166 \tab 906 \tab 388 \tab 15 \tab 20 \tab 111 \tab 2 \tab 5 \tab - \tab - \tab 1613 \cr
   1973 \tab 192 \tab 857 \tab 313 \tab 15 \tab 22 \tab 92 \tab 8 \tab 5 \tab - \tab - \tab 1504 \cr
   1974 \tab 193 \tab 784 \tab 367 \tab 8 \tab 25 \tab 104 \tab - \tab 3 \tab - \tab - \tab 1484 \cr
   1975 \tab 196 \tab 783 \tab 370 \tab 7 \tab 20 \tab 106 \tab - \tab 8 \tab - \tab - \tab 1490 \cr
   1976 \tab 188 \tab 763 \tab 403 \tab 14 \tab 21 \tab 106 \tab - \tab 4 \tab - \tab - \tab 1499 \cr
   1977 \tab 203 \tab 789 \tab 369 \tab 13 \tab 27 \tab 126 \tab - \tab 3 \tab - \tab - \tab 1530 \cr
   1978 \tab 182 \tab 829 \tab 361 \tab 13 \tab 22 \tab 123 \tab - \tab 2 \tab - \tab - \tab 1532 \cr
   1980 \tab 230 \tab 751 \tab 339 \tab 11 \tab 19 \tab 116 \tab - \tab 2 \tab - \tab - \tab 1468 \cr
   1982 \tab 233 \tab 948 \tab 440 \tab 18 \tab 38 \tab 179 \tab - \tab 4 \tab - \tab - \tab 1860 \cr
   1983 \tab 253 \tab 826 \tab 354 \tab 17 \tab 25 \tab 120 \tab - \tab 4 \tab - \tab - \tab 1599 \cr
   1984 \tab 205 \tab 773 \tab 346 \tab 7 \tab 6 \tab 127 \tab - \tab 9 \tab - \tab - \tab 1473 \cr
   1985 \tab 241 \tab 777 \tab 358 \tab 12 \tab 34 \tab 106 \tab - \tab 6 \tab - \tab - \tab 1534 \cr
   1986 \tab 245 \tab 738 \tab 329 \tab 15 \tab 31 \tab 109 \tab - \tab 3 \tab - \tab - \tab 1470 \cr
   1987 \tab 256 \tab 880 \tab 454 \tab 13 \tab 44 \tab 164 \tab - \tab 8 \tab - \tab - \tab 1819 \cr
   1988 \tab 229 \tab 747 \tab 349 \tab 14 \tab 23 \tab 114 \tab - \tab 5 \tab - \tab - \tab 1481 \cr
   1989 \tab 216 \tab 798 \tab 350 \tab 13 \tab 23 \tab 134 \tab - \tab 3 \tab - \tab - \tab 1537 \cr
   1990 \tab 216 \tab 651 \tab 342 \tab 12 \tab 22 \tab 126 \tab - \tab 3 \tab - \tab - \tab 1372 \cr
   1991 \tab 252 \tab 736 \tab 348 \tab 15 \tab 29 \tab 127 \tab - \tab 10 \tab - \tab - \tab 1517 \cr
   1993 \tab 252 \tab 795 \tab 393 \tab 12 \tab 34 \tab 116 \tab - \tab 4 \tab - \tab - \tab 1606 \cr
   1994 \tab 243 \tab 798 \tab 317 \tab 16 \tab 26 \tab 120 \tab - \tab 2 \tab 1470 \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab 2817 \cr
   2002 \tab 345 \tab 933 \tab 432 \tab 21 \tab 36 \tab 151 \tab - \tab 3 \tab 844 \tab - \tab 2765 \cr
   2004 \tab 377 \tab 1151 \tab 637 \tab 35 \tab 58 \tab 237 \tab - \tab 5 \tab 312 \tab - \tab 2812 \cr
   2006 \tab 475 \tab 1303 \tab 754 \tab 43 \tab 49 \tab 321 \tab - \tab 47 \tab 1518 \tab - \tab 4510 \cr
   2008 \tab 330 \tab 907 \tab 506 \tab 31 \tab 44 \tab 203 \tab - \tab 2 \tab - \tab - \tab 2023 \cr
   2010 \tab 353 \tab 843 \tab 550 \tab 34 \tab 51 \tab 206 \tab - \tab 7 \tab - \tab - \tab 2044 \cr
   2012 \tab 304 \tab 875 \tab 523 \tab 38 \tab 47 \tab 181 \tab - \tab 6 \tab - \tab - \tab 1974 \cr
   2014 \tab 403 \tab 1118 \tab 656 \tab 30 \tab 56 \tab 271 \tab - \tab 4 \tab - \tab - \tab 2538 \cr
   2016 \tab 501 \tab 1146 \tab 794 \tab 53 \tab 58 \tab 303 \tab - \tab 12 \tab - \tab - \tab 2867 \cr
   2018 \tab 371 \tab 995 \tab 641 \tab 38 \tab 50 \tab 245 \tab - \tab 8 \tab - \tab - \tab 2348 \cr
   2021 \tab 679 \tab 1625 \tab 1013 \tab 119 \tab 88 \tab 421 \tab - \tab - \tab 71 \tab 16 \tab 4032 \cr
   2022 \tab 540 \tab 1543 \tab 936 \tab 40 \tab 88 \tab 369 \tab - \tab 13 \tab 8 \tab 7 \tab 3544 \cr
   Total \tab 9069 \tab 28368 \tab 14732 \tab 742 \tab 1136 \tab 5334 \tab 10 \tab 200 \tab 12776 \tab 23 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} far below average
\item \code{2} below average
\item \code{3} average
\item \code{4} above average
\item \code{5} far above average
\item \code{7} lived in institution
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
