% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_03.R
\name{indstatus}
\alias{indstatus}
\title{Indus10 coding status}
\source{
General Social Survey https://gss.norc.org
}
\description{
indstatus
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab found, coded \tab found, iap \tab found, no answer \tab not found, coded \tab not found, iap \tab not found, no answer \tab not found, no code assigned \tab iap \tab Total \cr
   1972 \tab 626 \tab 68 \tab 8 \tab 640 \tab 94 \tab 17 \tab 160 \tab - \tab 1613 \cr
   1973 \tab 1092 \tab 117 \tab 25 \tab 156 \tab 30 \tab 11 \tab 73 \tab - \tab 1504 \cr
   1974 \tab 1037 \tab 108 \tab 72 \tab 209 \tab 20 \tab 1 \tab 37 \tab - \tab 1484 \cr
   1975 \tab 874 \tab 94 \tab 8 \tab 380 \tab 40 \tab 5 \tab 89 \tab - \tab 1490 \cr
   1976 \tab 1114 \tab 113 \tab 3 \tab 182 \tab 32 \tab 1 \tab 54 \tab - \tab 1499 \cr
   1977 \tab 761 \tab 54 \tab 2 \tab 534 \tab 59 \tab - \tab 120 \tab - \tab 1530 \cr
   1978 \tab 1398 \tab 114 \tab 6 \tab 1 \tab 13 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1194 \tab 96 \tab 3 \tab 123 \tab 23 \tab - \tab 29 \tab - \tab 1468 \cr
   1982 \tab 1565 \tab 95 \tab 36 \tab 113 \tab 27 \tab 5 \tab 19 \tab - \tab 1860 \cr
   1983 \tab 1382 \tab 85 \tab 8 \tab 88 \tab 18 \tab - \tab 18 \tab - \tab 1599 \cr
   1984 \tab 1373 \tab 78 \tab 7 \tab 4 \tab 10 \tab - \tab 1 \tab - \tab 1473 \cr
   1985 \tab 1226 \tab 67 \tab 3 \tab 177 \tab 14 \tab - \tab 47 \tab - \tab 1534 \cr
   1986 \tab 1339 \tab 84 \tab 2 \tab 22 \tab 16 \tab 1 \tab 6 \tab - \tab 1470 \cr
   1987 \tab 1447 \tab 60 \tab 4 \tab 210 \tab 35 \tab 7 \tab 56 \tab - \tab 1819 \cr
   1988 \tab 1379 \tab 76 \tab 5 \tab 11 \tab 9 \tab - \tab 1 \tab - \tab 1481 \cr
   1989 \tab 1347 \tab 70 \tab 3 \tab 83 \tab 22 \tab 1 \tab 11 \tab - \tab 1537 \cr
   1990 \tab 1292 \tab 61 \tab 5 \tab 4 \tab 10 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1412 \tab 84 \tab 3 \tab 5 \tab 13 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1477 \tab 61 \tab 7 \tab 38 \tab 17 \tab 1 \tab 5 \tab - \tab 1606 \cr
   1994 \tab 2804 \tab 120 \tab 14 \tab 38 \tab 6 \tab 3 \tab 7 \tab - \tab 2992 \cr
   1996 \tab 2766 \tab 102 \tab 15 \tab 14 \tab 4 \tab 2 \tab 1 \tab - \tab 2904 \cr
   1998 \tab 2644 \tab 121 \tab 36 \tab 14 \tab 14 \tab 1 \tab 2 \tab - \tab 2832 \cr
   2000 \tab 2668 \tab 125 \tab 24 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2651 \tab 91 \tab 23 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2654 \tab 113 \tab 45 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4284 \tab 187 \tab 39 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1950 \tab 62 \tab 11 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1928 \tab 86 \tab 30 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 47684 \tab 2592 \tab 447 \tab 3046 \tab 526 \tab 56 \tab 736 \tab 17303 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} found, iap
\item \code{2} found, no answer
\item \code{3} found, coded
\item \code{4} not found, iap
\item \code{5} not found, no answer
\item \code{6} not found, coded
\item \code{7} not found, no code assigned
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
