% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_02.R
\name{industry}
\alias{industry}
\title{industry}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's industry code (1970)
}
\details{
Question 2c. What kind of place (do/did) you work for? d. What (do/did) they (make/do)?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab NA(m) \tab NA(n) \tab accounting, auditing, and bookkeeping services \tab advertising \tab agricultural chemicals \tab agricultural production \tab agricultural services, except horticultural \tab air transportation \tab aircraft and parts \tab alcoholic beverages \tab apparel and accessories \tab apparel and accessories stores, except shoe stores \tab automobile repair and related services \tab automobile services, except repair \tab bakery products \tab banking \tab beauty shops \tab beverage industries \tab blast furnaces, steel works, rolling and finishing mills \tab bowling alleys, billiard and pool parlors \tab business management and consulting services \tab business services, n.e.c. \tab canning and preserving fruits, vegetables, and sea foods \tab cement concrete, gypsum, and  plaster products \tab coal mining \tab colleges and universities \tab computer programming services \tab confectionery and related products \tab construction and material handling machines \tab convalescent institutions \tab credit agencies \tab crude petroleum and natural gas extractions \tab cutlery, hand tools, and other hardware \tab cycles and miscellaneous transportation equipment \tab dairy product stores \tab dairy products \tab department and mail order establishments \tab detective and protective services \tab direct selling establishments \tab dressmaking shops \tab drug stores \tab drugs and medicines \tab drugs, chemicals, and allied products \tab dry goods and apparel \tab eating and drinking places \tab educational services, n.e.c. \tab electric light and power \tab electric-gas utilities \tab electrical goods \tab electrical machinery, equipment, and supplies, n.e.c. \tab electrical repair shops \tab electronic computing equipment \tab elementary and secondary schools \tab employment and temporary help agencies \tab engineering and architectural services \tab engines and turbines \tab fabricated structural metal products \tab farm and garden supply stores \tab farm machinery and equipment \tab farm products-raw materials \tab federal public administration \tab food and related products \tab food stores, n.e.c. \tab footwear, except rubber \tab forestry \tab furniture and home furnishing stores \tab furniture fixtures \tab gasoline service stations \tab general building contractors \tab general contractors, except building \tab glass and glass products \tab grain-mill products \tab grocery stores \tab hardware and farm equipment stores \tab hardware, plumbing, and heating supplies \tab health services, n.e.c. \tab horticultural services \tab hospitals \tab hotels and motels \tab household appliances \tab household appliances, tv, and radio stores \tab industrial chemicals \tab insurance \tab knitting mills \tab laundering, cleaning, and other garment services \tab legal services \tab libraries \tab limited price variety stores \tab local public administration \tab lodging places, except hotels and motels \tab logging \tab lumber and building material retailing \tab lumber and construction materials \tab machinery equipment and supplies \tab machinery, except electrical, n.e.c. \tab meat products \tab metal mining \tab metals and minerals, n.e.c. \tab metalworking machinery \tab misc. general merchandise stores \tab miscellaneous chemicals \tab miscellaneous entertainment and recreation services \tab miscellaneous fabricated metal products \tab miscellaneous fabricated textile products \tab miscellaneous food preparation and kindred products \tab miscellaneous manufacturing industries \tab miscellaneous nonmetallic mineral and stone products \tab miscellaneous paper and pulp products \tab miscellaneous personal services \tab miscellaneous plastic products \tab miscellaneous professional and related services \tab miscellaneous repair services \tab miscellaneous retail stores \tab miscellaneous vehicle dealers \tab mobile dwellings and campers \tab motor vehicle dealers \tab motor vehicles and equipment \tab motor vehicles and motor vehicle equipment \tab museums, art galleries, and zoos \tab newspaper publishing and printing \tab nonmetallic mining and quarrying, except fuel \tab nonprofit membership organizations \tab not specified chemicals and allied products \tab not specified construction \tab not specified electrical machinery, equipment, and supplies \tab not specified food industries \tab not specified machinery \tab not specified manufacturing industries \tab not specified metal industries \tab not specified retail trade \tab not specified wholesale trade \tab office and accounting machines \tab offices of dentists \tab offices of physicians \tab optical and health services supplies \tab ordinance \tab other primary iron and steel industries \tab other primary nonferrous industries \tab paints, varnishes, and related products \tab paper and its products \tab paperboard containers and boxes \tab petroleum products \tab petroleum refining \tab photographic equipment and supplies \tab pipe lines, except natural gas \tab postal service \tab pottery and related products \tab primary aluminum industries \tab printing, publishing, and allied industries, except newspapers \tab private households \tab pulp, paper and paperboard mills \tab radio broadcasting and television \tab radio, t.v., and communication equipment \tab railroad locomotives and equipment \tab railroads and railway express \tab real estate, incl. real estate-insurance-law offices \tab religious organizations \tab retail bakeries \tab retail florists \tab rubber products \tab sanitary services \tab sawmills, planing mills, and  mill work \tab scientific and controlling instruments \tab scrap and waste materials \tab security, commodity brokerage, and investment companies \tab services incidental to transportation \tab services to dwellings and other buildings \tab ship and boat building and repairing \tab soaps and cosmetics \tab special trade contractors \tab state public administration \tab street railways and bus lines \tab structural clay products \tab tanned, curried, and finished leather \tab telegraph and miscellaneous communication services \tab telephone (wire and radio) \tab theaters and motion pictures \tab tire, battery, and accessory dealers \tab trucking service \tab warehousing and storage \tab water supply \tab water transportation \tab welfare services \tab wholesalers, n.e.c. \tab yarn, thread, and fabric mills \tab commercial research, development, and testing labs \tab fisheries \tab floor coverings, except hard surface \tab fuel and ice dealers \tab gas and steam supply systems \tab jewelry stores \tab leather products, except footwear \tab liquor stores \tab metal stamping \tab miscellaneous textile mill products \tab offices of chiropractors \tab offices of health practitioners,n.e.c. \tab plastics, synthetics and resins, except fibers \tab residential welfare facilities \tab screw machine products \tab taxicab service \tab tobacco manufactures \tab watches, clocks, and clockwork-operated devices \tab miscellaneous wood products \tab other and not specified utilities \tab vending machine operators \tab barber shops \tab dyeing and finishing textiles, except wool and knit goods \tab shoe stores \tab synthetic fibers \tab not specified educational services \tab not specified electrical and hardware products \tab shoe repair shops \tab miscellaneous petroleum and coal products \tab not specified professional equipment \tab NA(y) \tab NA(i) \tab Total \cr
   1972 \tab 162 \tab 31 \tab 2 \tab 5 \tab 3 \tab 53 \tab 4 \tab 5 \tab 14 \tab 1 \tab 25 \tab 11 \tab 2 \tab 4 \tab 2 \tab 28 \tab 10 \tab 5 \tab 15 \tab 1 \tab 8 \tab 6 \tab 5 \tab 1 \tab 2 \tab 23 \tab 1 \tab 1 \tab 2 \tab 8 \tab 4 \tab 5 \tab 5 \tab 1 \tab 1 \tab 1 \tab 24 \tab 1 \tab 3 \tab 2 \tab 3 \tab 3 \tab 2 \tab 1 \tab 52 \tab 8 \tab 9 \tab 1 \tab 1 \tab 13 \tab 1 \tab 7 \tab 98 \tab 1 \tab 3 \tab 4 \tab 11 \tab 2 \tab 1 \tab 5 \tab 48 \tab 13 \tab 3 \tab 5 \tab 1 \tab 5 \tab 7 \tab 11 \tab 37 \tab 16 \tab 1 \tab 4 \tab 20 \tab 3 \tab 2 \tab 10 \tab 2 \tab 53 \tab 10 \tab 3 \tab 5 \tab 4 \tab 30 \tab 4 \tab 11 \tab 4 \tab 3 \tab 5 \tab 25 \tab 3 \tab 1 \tab 7 \tab 1 \tab 8 \tab 8 \tab 8 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 5 \tab 11 \tab 2 \tab 4 \tab 5 \tab 2 \tab 4 \tab 3 \tab 4 \tab 5 \tab 1 \tab 5 \tab 2 \tab 3 \tab 8 \tab 4 \tab 22 \tab 1 \tab 9 \tab 1 \tab 7 \tab 1 \tab 5 \tab 11 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 7 \tab 7 \tab 4 \tab 6 \tab 2 \tab 3 \tab 2 \tab 3 \tab 6 \tab 3 \tab 4 \tab 1 \tab 1 \tab 17 \tab 2 \tab 1 \tab 10 \tab 46 \tab 5 \tab 2 \tab 6 \tab 1 \tab 19 \tab 17 \tab 8 \tab 2 \tab 2 \tab 13 \tab 6 \tab 9 \tab 2 \tab 1 \tab 4 \tab 3 \tab 4 \tab 1 \tab 5 \tab 34 \tab 8 \tab 2 \tab 3 \tab 1 \tab 4 \tab 24 \tab 3 \tab 6 \tab 14 \tab 3 \tab 3 \tab 8 \tab 7 \tab 7 \tab 14 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 147 \tab 19 \tab 3 \tab 2 \tab - \tab 38 \tab 5 \tab 4 \tab 9 \tab 1 \tab 31 \tab 15 \tab 6 \tab - \tab 4 \tab 27 \tab 14 \tab 3 \tab 11 \tab 1 \tab 2 \tab 6 \tab 8 \tab 3 \tab 9 \tab 27 \tab - \tab 1 \tab 5 \tab 11 \tab 6 \tab 2 \tab 4 \tab - \tab - \tab - \tab 36 \tab 3 \tab 4 \tab 1 \tab 3 \tab 4 \tab 1 \tab 1 \tab 51 \tab 6 \tab 9 \tab 1 \tab 2 \tab 9 \tab 2 \tab 7 \tab 90 \tab - \tab 5 \tab 1 \tab 6 \tab - \tab 5 \tab 2 \tab 37 \tab 10 \tab 1 \tab 2 \tab - \tab 6 \tab 10 \tab 7 \tab 26 \tab 17 \tab 7 \tab 1 \tab 21 \tab 1 \tab 2 \tab 7 \tab 1 \tab 59 \tab 17 \tab 3 \tab 7 \tab 4 \tab 44 \tab 6 \tab 10 \tab 13 \tab 2 \tab - \tab 27 \tab 2 \tab 1 \tab 4 \tab 2 \tab 5 \tab 12 \tab 2 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 7 \tab 6 \tab 4 \tab 5 \tab 7 \tab 1 \tab 3 \tab 1 \tab 4 \tab 3 \tab 1 \tab 5 \tab - \tab 5 \tab 10 \tab 3 \tab 20 \tab - \tab 6 \tab 1 \tab 2 \tab 6 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 5 \tab 10 \tab 4 \tab 2 \tab 4 \tab 1 \tab 2 \tab 2 \tab 6 \tab 1 \tab 2 \tab 5 \tab - \tab 12 \tab 2 \tab 3 \tab 14 \tab 31 \tab 3 \tab 1 \tab 9 \tab - \tab 15 \tab 15 \tab 6 \tab 3 \tab 1 \tab 6 \tab 5 \tab 4 \tab - \tab 2 \tab 2 \tab - \tab 5 \tab 6 \tab 4 \tab 32 \tab 7 \tab - \tab 3 \tab - \tab 3 \tab 18 \tab 7 \tab 1 \tab 13 \tab 1 \tab 2 \tab 4 \tab 13 \tab 1 \tab 8 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 128 \tab 12 \tab 1 \tab 2 \tab - \tab 36 \tab - \tab 5 \tab 9 \tab - \tab 26 \tab 15 \tab 2 \tab - \tab 4 \tab 21 \tab 8 \tab 5 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 8 \tab 20 \tab 1 \tab 1 \tab 5 \tab 16 \tab 5 \tab 4 \tab 5 \tab - \tab 2 \tab 5 \tab 24 \tab - \tab 4 \tab - \tab 5 \tab 3 \tab - \tab 1 \tab 39 \tab 6 \tab 5 \tab 3 \tab 2 \tab 23 \tab 1 \tab 7 \tab 113 \tab 3 \tab 2 \tab 2 \tab 2 \tab - \tab 4 \tab 1 \tab 59 \tab 7 \tab 1 \tab 2 \tab - \tab 3 \tab 7 \tab 3 \tab 36 \tab 10 \tab 7 \tab 1 \tab 30 \tab 3 \tab - \tab 9 \tab 1 \tab 55 \tab 13 \tab 8 \tab 1 \tab 7 \tab 26 \tab 3 \tab 9 \tab 8 \tab 4 \tab 3 \tab 27 \tab 5 \tab 2 \tab 2 \tab - \tab 2 \tab 11 \tab 2 \tab 6 \tab 2 \tab 2 \tab 6 \tab 1 \tab 15 \tab 3 \tab 1 \tab 12 \tab 16 \tab - \tab 4 \tab 23 \tab 1 \tab 3 \tab 3 \tab 7 \tab 3 \tab 1 \tab 5 \tab 2 \tab 32 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 5 \tab 5 \tab 14 \tab 1 \tab - \tab 8 \tab - \tab 2 \tab 1 \tab 4 \tab 2 \tab 3 \tab 6 \tab 3 \tab 10 \tab - \tab 5 \tab 17 \tab 32 \tab 4 \tab - \tab 6 \tab 2 \tab 21 \tab 8 \tab 8 \tab 1 \tab 1 \tab 8 \tab 7 \tab 4 \tab 3 \tab - \tab 5 \tab 2 \tab 3 \tab 2 \tab 2 \tab 32 \tab 24 \tab 6 \tab - \tab - \tab - \tab 22 \tab 1 \tab 1 \tab 9 \tab 1 \tab 3 \tab - \tab 3 \tab 6 \tab 12 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 5 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 3 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 134 \tab 12 \tab 4 \tab 3 \tab - \tab 33 \tab 2 \tab 2 \tab 6 \tab 1 \tab 42 \tab 8 \tab 8 \tab 1 \tab - \tab 20 \tab 11 \tab 4 \tab 11 \tab - \tab 6 \tab 10 \tab 4 \tab 6 \tab 9 \tab 14 \tab 2 \tab - \tab 2 \tab 19 \tab 7 \tab 3 \tab 2 \tab 1 \tab - \tab 2 \tab 45 \tab - \tab 2 \tab - \tab 9 \tab 4 \tab - \tab 1 \tab 68 \tab 1 \tab 6 \tab - \tab - \tab 17 \tab 2 \tab 5 \tab 79 \tab 1 \tab 7 \tab 1 \tab 4 \tab 4 \tab 5 \tab 2 \tab 46 \tab 5 \tab 1 \tab 2 \tab 2 \tab 7 \tab 10 \tab 6 \tab 32 \tab 9 \tab 4 \tab 2 \tab 22 \tab 1 \tab 1 \tab 4 \tab 1 \tab 68 \tab 12 \tab 5 \tab 2 \tab 7 \tab 30 \tab 13 \tab 9 \tab 7 \tab 3 \tab 2 \tab 19 \tab 1 \tab 2 \tab 3 \tab 1 \tab 4 \tab 19 \tab 9 \tab 3 \tab - \tab 1 \tab 3 \tab - \tab 9 \tab 3 \tab 4 \tab - \tab 9 \tab 3 \tab - \tab - \tab 3 \tab 11 \tab 6 \tab 10 \tab - \tab 4 \tab 10 \tab 3 \tab 22 \tab 1 \tab 8 \tab 1 \tab 7 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 5 \tab 8 \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 19 \tab - \tab 2 \tab 8 \tab 38 \tab 2 \tab 1 \tab 8 \tab 1 \tab 10 \tab 10 \tab 14 \tab 2 \tab - \tab 9 \tab 5 \tab 4 \tab 3 \tab 3 \tab 5 \tab 1 \tab 5 \tab 3 \tab 3 \tab 36 \tab 12 \tab 6 \tab - \tab - \tab 1 \tab 14 \tab 5 \tab 3 \tab 15 \tab 4 \tab 3 \tab 1 \tab 5 \tab 5 \tab 5 \tab 1 \tab 1 \tab - \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 145 \tab 2 \tab 3 \tab 3 \tab 1 \tab 48 \tab 5 \tab 6 \tab 10 \tab 1 \tab 37 \tab 10 \tab 9 \tab 1 \tab 1 \tab 19 \tab 6 \tab 3 \tab 13 \tab - \tab 1 \tab 8 \tab 4 \tab - \tab 16 \tab 27 \tab 1 \tab - \tab 7 \tab 19 \tab 1 \tab 6 \tab 1 \tab - \tab 1 \tab 6 \tab 32 \tab 1 \tab 11 \tab 1 \tab 6 \tab 2 \tab 4 \tab - \tab 58 \tab - \tab 11 \tab 2 \tab 2 \tab 15 \tab 2 \tab 12 \tab 81 \tab 2 \tab 5 \tab 1 \tab 8 \tab - \tab 4 \tab - \tab 34 \tab 9 \tab 1 \tab 2 \tab 2 \tab 5 \tab 10 \tab 6 \tab 18 \tab 8 \tab 4 \tab 1 \tab 31 \tab 3 \tab 2 \tab 8 \tab 2 \tab 53 \tab 12 \tab 1 \tab 3 \tab 2 \tab 31 \tab 1 \tab 11 \tab 10 \tab - \tab 2 \tab 40 \tab 4 \tab 2 \tab 4 \tab 2 \tab 2 \tab 14 \tab 6 \tab 6 \tab - \tab 5 \tab 3 \tab - \tab 4 \tab 2 \tab 4 \tab 3 \tab 9 \tab 3 \tab 5 \tab 3 \tab 4 \tab 4 \tab 5 \tab 10 \tab 1 \tab - \tab 8 \tab 1 \tab 24 \tab - \tab 10 \tab - \tab 8 \tab 5 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 5 \tab 2 \tab 12 \tab 7 \tab 5 \tab 3 \tab 4 \tab 1 \tab - \tab 6 \tab 1 \tab 2 \tab 3 \tab - \tab 9 \tab 1 \tab 3 \tab 17 \tab 30 \tab 4 \tab 3 \tab 4 \tab - \tab 13 \tab 13 \tab 11 \tab 4 \tab 1 \tab 10 \tab 3 \tab 4 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 21 \tab 6 \tab 2 \tab 1 \tab 1 \tab - \tab 23 \tab 4 \tab 3 \tab 11 \tab 4 \tab 2 \tab 5 \tab 13 \tab 2 \tab 9 \tab 2 \tab 2 \tab - \tab 3 \tab 3 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 113 \tab 1 \tab 5 \tab 2 \tab 1 \tab 36 \tab 4 \tab 4 \tab 11 \tab - \tab 27 \tab 15 \tab 10 \tab 1 \tab 4 \tab 25 \tab 13 \tab 6 \tab 13 \tab - \tab 3 \tab 6 \tab 13 \tab 2 \tab 10 \tab 20 \tab 2 \tab 3 \tab 9 \tab 14 \tab 7 \tab 8 \tab 4 \tab 1 \tab 1 \tab 6 \tab 35 \tab 3 \tab 2 \tab - \tab 5 \tab 4 \tab 2 \tab - \tab 64 \tab 2 \tab 9 \tab - \tab 1 \tab 19 \tab 1 \tab 1 \tab 101 \tab 3 \tab 6 \tab - \tab 6 \tab - \tab 5 \tab 1 \tab 45 \tab 5 \tab 1 \tab 3 \tab 1 \tab 4 \tab 13 \tab 4 \tab 27 \tab 20 \tab 4 \tab 3 \tab 28 \tab 1 \tab - \tab 6 \tab 2 \tab 63 \tab 12 \tab 4 \tab 2 \tab 2 \tab 19 \tab 8 \tab 6 \tab 4 \tab 2 \tab 6 \tab 19 \tab 2 \tab 3 \tab 3 \tab 5 \tab 3 \tab 17 \tab 5 \tab 7 \tab - \tab 5 \tab 5 \tab - \tab 7 \tab 3 \tab 6 \tab 1 \tab 11 \tab - \tab 6 \tab 4 \tab - \tab - \tab 7 \tab 12 \tab 3 \tab 2 \tab 12 \tab 3 \tab 21 \tab - \tab 6 \tab 1 \tab 6 \tab 1 \tab 11 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 6 \tab 8 \tab 4 \tab 3 \tab 4 \tab - \tab 1 \tab - \tab 6 \tab 3 \tab 3 \tab - \tab - \tab 12 \tab 3 \tab 5 \tab 13 \tab 28 \tab 5 \tab 6 \tab 8 \tab - \tab 7 \tab 17 \tab 13 \tab 2 \tab 4 \tab 6 \tab 4 \tab 7 \tab 4 \tab 1 \tab 3 \tab 3 \tab 1 \tab 5 \tab 3 \tab 33 \tab 13 \tab 1 \tab 2 \tab - \tab - \tab 20 \tab 8 \tab 2 \tab 24 \tab 3 \tab 5 \tab - \tab 6 \tab 2 \tab 4 \tab 2 \tab 3 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 4 \tab 3 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 127 \tab 1 \tab 3 \tab 3 \tab 1 \tab 26 \tab 2 \tab 1 \tab 14 \tab 2 \tab 39 \tab 12 \tab 9 \tab 1 \tab 4 \tab 34 \tab 14 \tab 1 \tab 18 \tab - \tab - \tab 8 \tab 8 \tab 2 \tab 14 \tab 23 \tab 2 \tab - \tab 7 \tab 12 \tab 5 \tab 6 \tab 3 \tab 2 \tab - \tab 3 \tab 31 \tab 2 \tab 6 \tab 2 \tab 9 \tab 8 \tab 2 \tab 1 \tab 73 \tab - \tab 10 \tab 1 \tab 1 \tab 17 \tab 2 \tab 3 \tab 90 \tab 2 \tab 3 \tab 2 \tab 4 \tab - \tab 3 \tab 2 \tab 47 \tab 6 \tab 3 \tab 2 \tab 5 \tab 3 \tab 8 \tab 6 \tab 20 \tab 10 \tab 3 \tab 2 \tab 27 \tab 3 \tab 2 \tab 6 \tab 6 \tab 60 \tab 14 \tab 6 \tab - \tab 2 \tab 29 \tab 4 \tab 10 \tab 9 \tab 1 \tab 6 \tab 34 \tab 1 \tab - \tab 1 \tab 5 \tab 3 \tab 8 \tab 4 \tab 8 \tab - \tab 7 \tab 3 \tab 1 \tab 4 \tab 9 \tab 2 \tab 2 \tab 8 \tab 2 \tab 1 \tab 3 \tab 6 \tab 3 \tab 6 \tab 11 \tab - \tab - \tab 5 \tab - \tab 17 \tab 1 \tab 6 \tab 1 \tab 7 \tab 1 \tab 1 \tab 5 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 7 \tab 12 \tab 9 \tab 3 \tab 9 \tab 1 \tab 2 \tab 1 \tab 1 \tab 6 \tab 6 \tab 3 \tab - \tab 4 \tab 2 \tab 4 \tab 18 \tab 30 \tab 4 \tab 1 \tab 5 \tab 1 \tab 10 \tab 18 \tab 9 \tab 1 \tab 1 \tab 8 \tab 3 \tab 7 \tab 1 \tab - \tab 8 \tab 2 \tab 4 \tab 4 \tab 1 \tab 25 \tab 16 \tab 5 \tab - \tab - \tab 1 \tab 27 \tab 5 \tab 1 \tab 7 \tab 3 \tab 2 \tab 4 \tab 10 \tab 5 \tab 12 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 5 \tab 3 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 119 \tab 4 \tab 6 \tab 2 \tab - \tab 35 \tab 7 \tab 7 \tab 8 \tab 2 \tab 28 \tab 11 \tab 6 \tab 1 \tab 3 \tab 19 \tab 8 \tab 2 \tab 13 \tab 2 \tab 5 \tab 2 \tab 2 \tab 2 \tab 18 \tab 14 \tab 4 \tab 1 \tab 4 \tab 16 \tab 3 \tab 7 \tab 2 \tab - \tab - \tab 3 \tab 24 \tab 1 \tab 8 \tab - \tab 9 \tab 3 \tab 1 \tab 2 \tab 67 \tab 12 \tab 2 \tab 3 \tab 3 \tab 13 \tab 1 \tab 9 \tab 76 \tab 2 \tab 4 \tab 1 \tab 15 \tab 1 \tab - \tab 2 \tab 37 \tab 12 \tab 3 \tab 4 \tab 1 \tab 5 \tab 4 \tab 6 \tab 13 \tab 9 \tab - \tab 1 \tab 21 \tab 4 \tab 2 \tab 5 \tab 3 \tab 56 \tab 6 \tab 3 \tab 5 \tab 2 \tab 28 \tab 5 \tab 8 \tab 11 \tab 4 \tab 2 \tab 32 \tab 3 \tab 3 \tab 2 \tab 1 \tab 2 \tab 17 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 10 \tab 14 \tab 2 \tab 2 \tab 13 \tab - \tab 7 \tab 9 \tab 5 \tab 11 \tab 3 \tab 12 \tab 3 \tab 1 \tab 11 \tab 1 \tab 23 \tab - \tab 10 \tab 1 \tab 9 \tab 6 \tab 5 \tab 5 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 4 \tab 2 \tab 4 \tab 11 \tab 2 \tab 2 \tab 5 \tab 5 \tab - \tab 1 \tab 4 \tab 1 \tab 5 \tab 4 \tab - \tab 8 \tab 1 \tab 5 \tab 15 \tab 29 \tab 1 \tab 9 \tab 3 \tab 2 \tab 6 \tab 15 \tab 8 \tab 1 \tab - \tab 4 \tab 1 \tab 4 \tab 1 \tab - \tab 7 \tab 4 \tab 1 \tab 5 \tab 2 \tab 39 \tab 28 \tab 3 \tab 1 \tab - \tab 1 \tab 14 \tab 5 \tab - \tab 18 \tab 1 \tab 1 \tab 2 \tab 7 \tab 4 \tab 6 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 2 \tab - \tab 7 \tab - \tab - \tab 3 \tab - \tab - \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 122 \tab 18 \tab 3 \tab 3 \tab - \tab 36 \tab 2 \tab 3 \tab 17 \tab - \tab 40 \tab 17 \tab 12 \tab 3 \tab 2 \tab 30 \tab 19 \tab 3 \tab 10 \tab 4 \tab 5 \tab 25 \tab 11 \tab - \tab 8 \tab 30 \tab 3 \tab 4 \tab 4 \tab 30 \tab 5 \tab 6 \tab 3 \tab - \tab - \tab 2 \tab 27 \tab 6 \tab 2 \tab - \tab 8 \tab 2 \tab 3 \tab 5 \tab 66 \tab 9 \tab 12 \tab 2 \tab 3 \tab 13 \tab - \tab 3 \tab 116 \tab 3 \tab 7 \tab 5 \tab 5 \tab 2 \tab 3 \tab 3 \tab 35 \tab 13 \tab 2 \tab 1 \tab 4 \tab 8 \tab 9 \tab 8 \tab 19 \tab 19 \tab 3 \tab 2 \tab 27 \tab 2 \tab - \tab 13 \tab 2 \tab 83 \tab 11 \tab 4 \tab 3 \tab 3 \tab 32 \tab 3 \tab 7 \tab 12 \tab 1 \tab 4 \tab 24 \tab 11 \tab 4 \tab 6 \tab - \tab 7 \tab 19 \tab 10 \tab 7 \tab - \tab 5 \tab 1 \tab 2 \tab 13 \tab 7 \tab - \tab 12 \tab 11 \tab 1 \tab 7 \tab 11 \tab 3 \tab 2 \tab 8 \tab 12 \tab 3 \tab 3 \tab 9 \tab 2 \tab 21 \tab 3 \tab 5 \tab 5 \tab 10 \tab 6 \tab 7 \tab 1 \tab 1 \tab 3 \tab 6 \tab 4 \tab 4 \tab 4 \tab 1 \tab 6 \tab 15 \tab 3 \tab 4 \tab 10 \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab 3 \tab 10 \tab 1 \tab 15 \tab 1 \tab 2 \tab 18 \tab 65 \tab 1 \tab 3 \tab 6 \tab 1 \tab 13 \tab 16 \tab 8 \tab 2 \tab 2 \tab 3 \tab 9 \tab 2 \tab 2 \tab 1 \tab 7 \tab - \tab 9 \tab 6 \tab 3 \tab 53 \tab 15 \tab 4 \tab 1 \tab - \tab 2 \tab 23 \tab 3 \tab 5 \tab 23 \tab - \tab 3 \tab 2 \tab 21 \tab 6 \tab 8 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 4 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 103 \tab 10 \tab 8 \tab 6 \tab 2 \tab 41 \tab 7 \tab 5 \tab 5 \tab 1 \tab 25 \tab 12 \tab 12 \tab 5 \tab 1 \tab 25 \tab 9 \tab 2 \tab 22 \tab 1 \tab 1 \tab 12 \tab 1 \tab 3 \tab 7 \tab 38 \tab 5 \tab - \tab 6 \tab 17 \tab 4 \tab 6 \tab - \tab 1 \tab 3 \tab 4 \tab 24 \tab 2 \tab 4 \tab 1 \tab 5 \tab 1 \tab 4 \tab 2 \tab 80 \tab 6 \tab 3 \tab 3 \tab 3 \tab 12 \tab 2 \tab 7 \tab 96 \tab 2 \tab 8 \tab 5 \tab 16 \tab 1 \tab 4 \tab - \tab 48 \tab 14 \tab 2 \tab 6 \tab - \tab 5 \tab 7 \tab 6 \tab 26 \tab 2 \tab 4 \tab - \tab 29 \tab 5 \tab 2 \tab 14 \tab 2 \tab 79 \tab 11 \tab 3 \tab 5 \tab 4 \tab 32 \tab 4 \tab 7 \tab 9 \tab 1 \tab 2 \tab 50 \tab - \tab 1 \tab 2 \tab - \tab 6 \tab 10 \tab 13 \tab 8 \tab - \tab 2 \tab 6 \tab 1 \tab 13 \tab 9 \tab 1 \tab 13 \tab 5 \tab 4 \tab 1 \tab 6 \tab 1 \tab 4 \tab 7 \tab 12 \tab - \tab - \tab 5 \tab 2 \tab 11 \tab 3 \tab 3 \tab 1 \tab 6 \tab 2 \tab - \tab 4 \tab 2 \tab - \tab 6 \tab 2 \tab 2 \tab 1 \tab 3 \tab 12 \tab 11 \tab 10 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 4 \tab 2 \tab 4 \tab - \tab - \tab 6 \tab - \tab 3 \tab 10 \tab 24 \tab 4 \tab 4 \tab 7 \tab - \tab 6 \tab 21 \tab 7 \tab 2 \tab 4 \tab 3 \tab 3 \tab 1 \tab 2 \tab 2 \tab 10 \tab 1 \tab 2 \tab 1 \tab 2 \tab 49 \tab 24 \tab 1 \tab - \tab - \tab 2 \tab 21 \tab 2 \tab 4 \tab 10 \tab 1 \tab - \tab - \tab - \tab 5 \tab 8 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 3 \tab 4 \tab - \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 4 \tab 3 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1599 \cr
   1984 \tab 88 \tab 6 \tab 8 \tab 7 \tab - \tab 23 \tab 7 \tab 7 \tab 13 \tab - \tab 13 \tab 10 \tab 10 \tab 2 \tab 6 \tab 25 \tab 12 \tab 1 \tab 6 \tab - \tab 7 \tab 14 \tab 7 \tab 1 \tab 3 \tab 23 \tab 1 \tab - \tab - \tab 17 \tab 9 \tab 5 \tab 4 \tab - \tab - \tab 5 \tab 25 \tab - \tab 2 \tab - \tab 2 \tab 8 \tab 1 \tab - \tab 76 \tab 7 \tab 9 \tab 2 \tab 3 \tab 12 \tab - \tab 9 \tab 96 \tab 3 \tab 4 \tab - \tab 3 \tab 3 \tab - \tab 3 \tab 39 \tab 10 \tab 6 \tab 7 \tab - \tab 6 \tab 7 \tab 7 \tab 36 \tab 3 \tab 3 \tab 1 \tab 30 \tab 3 \tab 2 \tab 23 \tab 1 \tab 63 \tab 14 \tab 5 \tab - \tab 4 \tab 17 \tab 4 \tab 17 \tab 8 \tab - \tab 2 \tab 33 \tab 1 \tab 3 \tab 2 \tab 1 \tab 2 \tab 5 \tab 6 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 8 \tab 12 \tab - \tab 7 \tab 42 \tab 2 \tab 5 \tab 14 \tab 2 \tab 3 \tab 1 \tab 23 \tab - \tab 1 \tab 9 \tab 4 \tab 20 \tab - \tab 6 \tab 3 \tab 5 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 5 \tab 6 \tab 5 \tab - \tab 16 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 4 \tab - \tab 6 \tab 1 \tab 2 \tab 10 \tab 49 \tab - \tab - \tab 6 \tab 1 \tab 6 \tab 20 \tab 2 \tab - \tab 3 \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab 6 \tab 4 \tab 4 \tab 2 \tab 3 \tab 14 \tab 18 \tab 3 \tab 1 \tab - \tab 13 \tab 6 \tab 3 \tab 1 \tab 13 \tab - \tab - \tab 3 \tab 3 \tab 10 \tab 18 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 81 \tab - \tab 4 \tab 3 \tab - \tab 39 \tab 5 \tab 4 \tab 8 \tab 1 \tab 29 \tab 12 \tab 7 \tab 3 \tab 3 \tab 28 \tab 14 \tab 4 \tab 10 \tab 2 \tab 8 \tab 14 \tab 6 \tab 2 \tab - \tab 35 \tab 5 \tab - \tab 1 \tab 18 \tab 5 \tab 6 \tab 2 \tab 1 \tab - \tab 3 \tab 32 \tab 4 \tab 3 \tab - \tab 3 \tab 3 \tab 4 \tab 1 \tab 79 \tab 3 \tab 10 \tab 1 \tab 2 \tab 10 \tab 1 \tab 9 \tab 85 \tab 1 \tab 9 \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab 36 \tab 13 \tab 5 \tab 1 \tab 2 \tab 5 \tab 7 \tab 5 \tab 36 \tab 14 \tab 6 \tab - \tab 12 \tab 4 \tab 2 \tab 17 \tab 6 \tab 75 \tab 12 \tab - \tab 8 \tab 4 \tab 44 \tab - \tab 5 \tab 8 \tab 2 \tab 3 \tab 31 \tab 5 \tab 1 \tab 3 \tab 2 \tab 9 \tab 13 \tab 6 \tab 2 \tab - \tab 4 \tab 5 \tab 1 \tab 9 \tab 5 \tab 1 \tab - \tab 8 \tab 2 \tab 3 \tab 7 \tab 5 \tab 6 \tab 5 \tab 14 \tab - \tab 2 \tab 9 \tab 1 \tab 13 \tab 2 \tab 4 \tab - \tab 5 \tab 4 \tab 3 \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 2 \tab 9 \tab 12 \tab 4 \tab 2 \tab 3 \tab 3 \tab - \tab - \tab 2 \tab 3 \tab 5 \tab 2 \tab - \tab 10 \tab 1 \tab - \tab 20 \tab 28 \tab 2 \tab 6 \tab 3 \tab 1 \tab 12 \tab 26 \tab 10 \tab 1 \tab 3 \tab 4 \tab 4 \tab 2 \tab 3 \tab 6 \tab 3 \tab 4 \tab 3 \tab 2 \tab 2 \tab 46 \tab 18 \tab 2 \tab - \tab - \tab 2 \tab 27 \tab 5 \tab 4 \tab 14 \tab 2 \tab 1 \tab 3 \tab 5 \tab 4 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 6 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 100 \tab 5 \tab 10 \tab 4 \tab - \tab 36 \tab 6 \tab 4 \tab 16 \tab - \tab 22 \tab 7 \tab 12 \tab 2 \tab - \tab 31 \tab 16 \tab 6 \tab 9 \tab 2 \tab 2 \tab 19 \tab 2 \tab 2 \tab 1 \tab 29 \tab 7 \tab - \tab 3 \tab 34 \tab 1 \tab 7 \tab 3 \tab 1 \tab - \tab 1 \tab 32 \tab 4 \tab 3 \tab 1 \tab 6 \tab 1 \tab - \tab - \tab 59 \tab 5 \tab 6 \tab 3 \tab 2 \tab 9 \tab 1 \tab 10 \tab 82 \tab 5 \tab 7 \tab 1 \tab 4 \tab - \tab 2 \tab 2 \tab 36 \tab 8 \tab 2 \tab 3 \tab 3 \tab 3 \tab 8 \tab 7 \tab 24 \tab 8 \tab 3 \tab 2 \tab 26 \tab 2 \tab 4 \tab 11 \tab 5 \tab 68 \tab 13 \tab 4 \tab 3 \tab 2 \tab 25 \tab 1 \tab 13 \tab 6 \tab 3 \tab 6 \tab 35 \tab 2 \tab 2 \tab 1 \tab 3 \tab 4 \tab 8 \tab 3 \tab 2 \tab 1 \tab 4 \tab 3 \tab 1 \tab 13 \tab 3 \tab 1 \tab 4 \tab 10 \tab 1 \tab 3 \tab 3 \tab 2 \tab 2 \tab 5 \tab 14 \tab - \tab - \tab 7 \tab 1 \tab 24 \tab - \tab 6 \tab 1 \tab 5 \tab 2 \tab 2 \tab 5 \tab 2 \tab 4 \tab - \tab 2 \tab 2 \tab - \tab 4 \tab 12 \tab 10 \tab 3 \tab 6 \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab 4 \tab 2 \tab - \tab 8 \tab 2 \tab - \tab 16 \tab 25 \tab 4 \tab 2 \tab 5 \tab - \tab 6 \tab 23 \tab 4 \tab 3 \tab 1 \tab 1 \tab - \tab 5 \tab 3 \tab 4 \tab 6 \tab 1 \tab - \tab 1 \tab 6 \tab 32 \tab 21 \tab 4 \tab - \tab - \tab 2 \tab 10 \tab 1 \tab 4 \tab 13 \tab 5 \tab - \tab 2 \tab 4 \tab 9 \tab 10 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 4 \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 95 \tab 8 \tab 4 \tab 3 \tab - \tab 40 \tab 8 \tab 3 \tab 19 \tab - \tab 29 \tab 9 \tab 6 \tab 5 \tab 6 \tab 31 \tab 19 \tab 4 \tab 12 \tab - \tab 6 \tab 15 \tab - \tab 3 \tab 3 \tab 40 \tab 9 \tab 2 \tab 2 \tab 41 \tab 7 \tab 3 \tab 4 \tab 1 \tab - \tab 1 \tab 34 \tab 1 \tab 6 \tab - \tab 6 \tab 4 \tab - \tab - \tab 78 \tab 7 \tab 9 \tab 1 \tab 3 \tab 10 \tab - \tab 8 \tab 96 \tab 5 \tab 10 \tab 2 \tab 5 \tab 2 \tab 1 \tab 1 \tab 47 \tab 11 \tab 1 \tab 6 \tab - \tab 10 \tab 4 \tab 9 \tab 26 \tab 2 \tab 6 \tab - \tab 39 \tab 4 \tab - \tab 23 \tab 6 \tab 81 \tab 23 \tab 1 \tab 6 \tab - \tab 32 \tab 3 \tab 9 \tab 15 \tab 3 \tab 4 \tab 39 \tab 4 \tab 1 \tab 3 \tab 2 \tab 4 \tab 5 \tab 7 \tab 1 \tab 1 \tab 4 \tab 6 \tab 2 \tab 13 \tab 8 \tab 2 \tab 3 \tab 15 \tab 3 \tab 3 \tab 10 \tab - \tab 9 \tab 12 \tab 23 \tab - \tab - \tab 6 \tab 1 \tab 28 \tab 1 \tab 10 \tab 2 \tab 2 \tab 6 \tab 13 \tab 4 \tab - \tab 1 \tab 7 \tab 4 \tab 3 \tab 2 \tab 2 \tab 6 \tab 14 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab - \tab - \tab 13 \tab - \tab 1 \tab 15 \tab 59 \tab 3 \tab 10 \tab 5 \tab 1 \tab 12 \tab 26 \tab 6 \tab 3 \tab 2 \tab 3 \tab 3 \tab 6 \tab 2 \tab 2 \tab 6 \tab 5 \tab 2 \tab 4 \tab 2 \tab 62 \tab 21 \tab 7 \tab 2 \tab - \tab 5 \tab 23 \tab 2 \tab 7 \tab 14 \tab 2 \tab 2 \tab 1 \tab 8 \tab 12 \tab 10 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 9 \tab - \tab 1 \tab 3 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab 4 \tab 5 \tab - \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 85 \tab 3 \tab 7 \tab 5 \tab 2 \tab 24 \tab 2 \tab 7 \tab 9 \tab 2 \tab 25 \tab 15 \tab 13 \tab 2 \tab 3 \tab 22 \tab 10 \tab 9 \tab 1 \tab 2 \tab 12 \tab 11 \tab 1 \tab 4 \tab 3 \tab 29 \tab 3 \tab 2 \tab 3 \tab 18 \tab 10 \tab - \tab 1 \tab - \tab - \tab 1 \tab 32 \tab 3 \tab 1 \tab - \tab 7 \tab 4 \tab - \tab - \tab 61 \tab 1 \tab 9 \tab 3 \tab - \tab 9 \tab 1 \tab 3 \tab 77 \tab 5 \tab 8 \tab 1 \tab 10 \tab - \tab 3 \tab - \tab 41 \tab 10 \tab 5 \tab - \tab - \tab 5 \tab 11 \tab 6 \tab 25 \tab 4 \tab 2 \tab - \tab 26 \tab 4 \tab 3 \tab 15 \tab 2 \tab 73 \tab 11 \tab 3 \tab 3 \tab 4 \tab 22 \tab 2 \tab 5 \tab 8 \tab 1 \tab 4 \tab 33 \tab 2 \tab 4 \tab 1 \tab 2 \tab 3 \tab 11 \tab 5 \tab - \tab - \tab 3 \tab 11 \tab 1 \tab 14 \tab 1 \tab 2 \tab 2 \tab 11 \tab 2 \tab 2 \tab 8 \tab 5 \tab 6 \tab 4 \tab 20 \tab 3 \tab - \tab 9 \tab 3 \tab 16 \tab 1 \tab 4 \tab 3 \tab 9 \tab 3 \tab 3 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 4 \tab 12 \tab 7 \tab - \tab 5 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 13 \tab 1 \tab - \tab 12 \tab 38 \tab 4 \tab 4 \tab 6 \tab - \tab 14 \tab 15 \tab 10 \tab 3 \tab 2 \tab 4 \tab 1 \tab 2 \tab 3 \tab 2 \tab 3 \tab 2 \tab 13 \tab 2 \tab 1 \tab 50 \tab 17 \tab 6 \tab - \tab - \tab 1 \tab 24 \tab 7 \tab 2 \tab 11 \tab 6 \tab 1 \tab 2 \tab 6 \tab 2 \tab 3 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab 8 \tab 1 \tab 1 \tab 8 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 92 \tab 7 \tab 4 \tab 1 \tab 1 \tab 23 \tab 6 \tab 11 \tab 6 \tab - \tab 16 \tab 14 \tab 8 \tab 5 \tab 1 \tab 28 \tab 6 \tab 6 \tab 2 \tab 1 \tab 7 \tab 13 \tab 3 \tab - \tab 5 \tab 49 \tab 6 \tab - \tab 4 \tab 20 \tab 7 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 24 \tab 3 \tab 3 \tab - \tab 7 \tab 4 \tab 5 \tab 3 \tab 66 \tab 9 \tab 10 \tab 4 \tab 2 \tab 7 \tab 6 \tab 5 \tab 86 \tab 7 \tab 6 \tab 4 \tab 9 \tab 1 \tab - \tab - \tab 52 \tab 8 \tab - \tab 4 \tab - \tab 8 \tab 5 \tab 5 \tab 20 \tab 10 \tab 3 \tab 1 \tab 32 \tab 5 \tab 1 \tab 12 \tab 5 \tab 84 \tab 12 \tab 5 \tab 2 \tab 4 \tab 32 \tab 5 \tab 10 \tab 19 \tab 3 \tab 3 \tab 21 \tab 9 \tab 2 \tab 5 \tab 3 \tab - \tab 10 \tab 9 \tab 1 \tab - \tab 7 \tab 7 \tab 2 \tab 9 \tab 1 \tab 2 \tab 3 \tab 17 \tab 1 \tab 4 \tab 6 \tab 2 \tab 3 \tab 4 \tab 9 \tab - \tab 1 \tab 13 \tab 1 \tab 13 \tab - \tab 5 \tab - \tab 10 \tab 1 \tab 13 \tab 5 \tab 1 \tab 1 \tab 4 \tab 1 \tab 11 \tab 2 \tab 1 \tab 7 \tab 6 \tab 7 \tab 3 \tab 4 \tab - \tab 1 \tab 2 \tab 5 \tab - \tab 4 \tab 1 \tab - \tab 8 \tab 1 \tab 3 \tab 19 \tab 17 \tab 3 \tab 3 \tab 2 \tab - \tab 2 \tab 23 \tab 6 \tab 4 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab 7 \tab 8 \tab 8 \tab 2 \tab - \tab 36 \tab 15 \tab 6 \tab 1 \tab - \tab - \tab 19 \tab 5 \tab 5 \tab 16 \tab 2 \tab 2 \tab 2 \tab 6 \tab 6 \tab 8 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 5 \tab 4 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab 2 \tab 2 \tab 3 \tab - \tab 2 \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 71 \tab 7 \tab 7 \tab 2 \tab - \tab 27 \tab 5 \tab 4 \tab 11 \tab 1 \tab 19 \tab 11 \tab 10 \tab 2 \tab 3 \tab 24 \tab 9 \tab - \tab 8 \tab 1 \tab 6 \tab 17 \tab 1 \tab 2 \tab - \tab 35 \tab 4 \tab 1 \tab 3 \tab 14 \tab 4 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 21 \tab 1 \tab 2 \tab - \tab 6 \tab 3 \tab 2 \tab - \tab 61 \tab 8 \tab 4 \tab 3 \tab - \tab 11 \tab 3 \tab 4 \tab 97 \tab 8 \tab 6 \tab 1 \tab 9 \tab 1 \tab - \tab 1 \tab 48 \tab 12 \tab 2 \tab 5 \tab 3 \tab 1 \tab 5 \tab 2 \tab 13 \tab 16 \tab 5 \tab 2 \tab 19 \tab 1 \tab 1 \tab 13 \tab 3 \tab 55 \tab 9 \tab - \tab - \tab 4 \tab 28 \tab 8 \tab 6 \tab 12 \tab 3 \tab 1 \tab 18 \tab 1 \tab 2 \tab 2 \tab 3 \tab 6 \tab 4 \tab 8 \tab - \tab - \tab 4 \tab 2 \tab 1 \tab 10 \tab 7 \tab 1 \tab 2 \tab 10 \tab - \tab 3 \tab 4 \tab 4 \tab 7 \tab 6 \tab 18 \tab - \tab - \tab 9 \tab - \tab 21 \tab - \tab 4 \tab 1 \tab 6 \tab 3 \tab 7 \tab 4 \tab 1 \tab 3 \tab - \tab 1 \tab 9 \tab 2 \tab - \tab 5 \tab 6 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 4 \tab - \tab - \tab 9 \tab 2 \tab 2 \tab 14 \tab 25 \tab 2 \tab 7 \tab 2 \tab - \tab 4 \tab 24 \tab 5 \tab 3 \tab 4 \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab 5 \tab 4 \tab 8 \tab 1 \tab 1 \tab 53 \tab 17 \tab 3 \tab 1 \tab - \tab 3 \tab 15 \tab 3 \tab - \tab 19 \tab 3 \tab - \tab 2 \tab 17 \tab 5 \tab 8 \tab 5 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 1912 \tab 146 \tab 82 \tab 56 \tab 11 \tab 594 \tab 77 \tab 82 \tab 185 \tab 13 \tab 473 \tab 204 \tab 142 \tab 38 \tab 47 \tab 437 \tab 198 \tab 64 \tab 174 \tab 18 \tab 79 \tab 188 \tab 78 \tab 34 \tab 116 \tab 476 \tab 56 \tab 16 \tab 67 \tab 324 \tab 90 \tab 81 \tab 45 \tab 9 \tab 8 \tab 45 \tab 502 \tab 35 \tab 66 \tab 8 \tab 99 \tab 61 \tab 31 \tab 18 \tab 1098 \tab 90 \tab 133 \tab 33 \tab 30 \tab 219 \tab 26 \tab 109 \tab 1559 \tab 53 \tab 100 \tab 32 \tab 121 \tab 19 \tab 41 \tab 25 \tab 735 \tab 166 \tab 39 \tab 55 \tab 24 \tab 89 \tab 132 \tab 104 \tab 434 \tab 177 \tab 65 \tab 23 \tab 440 \tab 49 \tab 26 \tab 196 \tab 50 \tab 1128 \tab 212 \tab 58 \tab 55 \tab 59 \tab 501 \tab 74 \tab 153 \tab 163 \tab 36 \tab 55 \tab 507 \tab 56 \tab 34 \tab 51 \tab 33 \tab 70 \tab 191 \tab 106 \tab 56 \tab 5 \tab 59 \tab 67 \tab 19 \tab 163 \tab 104 \tab 35 \tab 75 \tab 207 \tab 27 \tab 61 \tab 115 \tab 51 \tab 82 \tab 84 \tab 217 \tab 18 \tab 23 \tab 145 \tab 32 \tab 348 \tab 14 \tab 106 \tab 23 \tab 104 \tab 49 \tab 82 \tab 61 \tab 16 \tab 26 \tab 37 \tab 18 \tab 46 \tab 27 \tab 40 \tab 107 \tab 174 \tab 76 \tab 40 \tab 85 \tab 30 \tab 24 \tab 23 \tab 57 \tab 27 \tab 58 \tab 46 \tab 5 \tab 179 \tab 20 \tab 41 \tab 246 \tab 594 \tab 51 \tab 62 \tab 91 \tab 11 \tab 176 \tab 307 \tab 135 \tab 37 \tab 32 \tab 91 \tab 57 \tab 68 \tab 36 \tab 27 \tab 88 \tab 45 \tab 74 \tab 50 \tab 41 \tab 647 \tab 280 \tab 61 \tab 16 \tab 2 \tab 40 \tab 330 \tab 69 \tab 49 \tab 244 \tab 41 \tab 30 \tab 41 \tab 134 \tab 91 \tab 146 \tab 27 \tab 12 \tab 8 \tab 18 \tab 33 \tab 24 \tab 14 \tab 23 \tab 19 \tab 48 \tab 11 \tab 18 \tab 40 \tab 24 \tab 10 \tab 28 \tab 21 \tab 19 \tab 31 \tab 27 \tab 11 \tab 27 \tab 10 \tab 23 \tab 14 \tab 9 \tab 9 \tab 5 \tab 8 \tab 1 \tab 42581 \tab 3544 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{17} agricultural production
\item \code{18} agricultural services, except horticultural
\item \code{19} horticultural services
\item \code{27} forestry
\item \code{28} fisheries
\item \code{47} metal mining
\item \code{48} coal mining
\item \code{49} crude petroleum and natural gas extractions
\item \code{57} nonmetallic mining and quarrying, except fuel
\item \code{67} general building contractors
\item \code{68} general contractors, except building
\item \code{69} special trade contractors
\item \code{77} not specified construction
\item \code{107} logging
\item \code{108} sawmills, planing mills, and mill work
\item \code{109} miscellaneous wood products
\item \code{118} furniture fixtures
\item \code{119} glass and glass products
\item \code{127} cement concrete, gypsum, and plaster products
\item \code{128} structural clay products
\item \code{137} pottery and related products
\item \code{138} miscellaneous nonmetallic mineral and stone products
\item \code{139} blast furnaces, steel works, rolling and finishing mills
\item \code{147} other primary iron and steel industries
\item \code{148} primary aluminum industries
\item \code{149} other primary nonferrous industries
\item \code{157} cutlery, hand tools, and other hardware
\item \code{158} fabricated structural metal products
\item \code{159} screw machine products
\item \code{167} metal stamping
\item \code{168} miscellaneous fabricated metal products
\item \code{169} not specified metal industries
\item \code{177} engines and turbines
\item \code{178} farm machinery and equipment
\item \code{179} construction and material handling machines
\item \code{187} metalworking machinery
\item \code{188} office and accounting machines
\item \code{189} electronic computing equipment
\item \code{197} machinery, except electrical, n.e.c.
\item \code{198} not specified machinery
\item \code{199} household appliances
\item \code{207} radio, t.v., and communication equipment
\item \code{208} electrical machinery, equipment, and supplies, n.e.c.
\item \code{209} not specified electrical machinery, equipment, and supplies
\item \code{219} motor vehicles and motor vehicle equipment
\item \code{227} aircraft and parts
\item \code{228} ship and boat building and repairing
\item \code{229} railroad locomotives and equipment
\item \code{237} mobile dwellings and campers
\item \code{238} cycles and miscellaneous transportation equipment
\item \code{239} scientific and controlling instruments
\item \code{247} optical and health services supplies
\item \code{248} photographic equipment and supplies
\item \code{249} watches, clocks, and clockwork-operated devices
\item \code{257} not specified professional equipment
\item \code{258} ordinance
\item \code{259} miscellaneous manufacturing industries
\item \code{268} meat products
\item \code{269} dairy products
\item \code{278} canning and preserving fruits, vegetables, and sea foods
\item \code{279} grain-mill products
\item \code{287} bakery products
\item \code{288} confectionery and related products
\item \code{289} beverage industries
\item \code{297} miscellaneous food preparation and kindred products
\item \code{298} not specified food industries
\item \code{299} tobacco manufactures
\item \code{307} knitting mills
\item \code{308} dyeing and finishing textiles, except wool and knit goods
\item \code{309} floor coverings, except hard surface
\item \code{317} yarn, thread, and fabric mills
\item \code{318} miscellaneous textile mill products
\item \code{319} apparel and accessories
\item \code{327} miscellaneous fabricated textile products
\item \code{328} pulp, paper and paperboard mills
\item \code{329} miscellaneous paper and pulp products
\item \code{337} paperboard containers and boxes
\item \code{338} newspaper publishing and printing
\item \code{339} printing, publishing, and allied industries, except newspapers
\item \code{347} industrial chemicals
\item \code{348} plastics, synthetics and resins, except fibers
\item \code{349} synthetic fibers
\item \code{357} drugs and medicines
\item \code{358} soaps and cosmetics
\item \code{359} paints, varnishes, and related products
\item \code{367} agricultural chemicals
\item \code{368} miscellaneous chemicals
\item \code{369} not specified chemicals and allied products
\item \code{377} petroleum refining
\item \code{378} miscellaneous petroleum and coal products
\item \code{379} rubber products
\item \code{387} miscellaneous plastic products
\item \code{388} tanned, curried, and finished leather
\item \code{389} footwear, except rubber
\item \code{397} leather products, except footwear
\item \code{398} not specified manufacturing industries
\item \code{407} railroads and railway express
\item \code{408} street railways and bus lines
\item \code{409} taxicab service
\item \code{417} trucking service
\item \code{418} warehousing and storage
\item \code{419} water transportation
\item \code{427} air transportation
\item \code{428} pipe lines, except natural gas
\item \code{429} services incidental to transportation
\item \code{447} radio broadcasting and television
\item \code{448} telephone (wire and radio)
\item \code{449} telegraph and miscellaneous communication services
\item \code{467} electric light and power
\item \code{468} electric-gas utilities
\item \code{469} gas and steam supply systems
\item \code{477} water supply
\item \code{478} sanitary services
\item \code{479} other and not specified utilities
\item \code{507} motor vehicles and equipment
\item \code{508} drugs, chemicals, and allied products
\item \code{509} dry goods and apparel
\item \code{527} food and related products
\item \code{528} farm products-raw materials
\item \code{529} electrical goods
\item \code{537} hardware, plumbing, and heating supplies
\item \code{538} not specified electrical and hardware products
\item \code{539} machinery equipment and supplies
\item \code{557} metals and minerals, n.e.c.
\item \code{558} petroleum products
\item \code{559} scrap and waste materials
\item \code{567} alcoholic beverages
\item \code{568} paper and its products
\item \code{569} lumber and construction materials
\item \code{587} wholesalers, n.e.c.
\item \code{588} not specified wholesale trade
\item \code{607} lumber and building material retailing
\item \code{608} hardware and farm equipment stores
\item \code{609} department and mail order establishments
\item \code{617} limited price variety stores
\item \code{618} vending machine operators
\item \code{619} direct selling establishments
\item \code{627} misc. general merchandise stores
\item \code{628} grocery stores
\item \code{629} dairy product stores
\item \code{637} retail bakeries
\item \code{638} food stores, n.e.c.
\item \code{639} motor vehicle dealers
\item \code{647} tire, battery, and accessory dealers
\item \code{648} gasoline service stations
\item \code{649} miscellaneous vehicle dealers
\item \code{657} apparel and accessories stores, except shoe stores
\item \code{658} shoe stores
\item \code{667} furniture and home furnishing stores
\item \code{668} household appliances, tv, and radio stores
\item \code{669} eating and drinking places
\item \code{677} drug stores
\item \code{678} liquor stores
\item \code{679} farm and garden supply stores
\item \code{687} jewelry stores
\item \code{688} fuel and ice dealers
\item \code{689} retail florists
\item \code{697} miscellaneous retail stores
\item \code{698} not specified retail trade
\item \code{707} banking
\item \code{708} credit agencies
\item \code{709} security, commodity brokerage, and investment companies
\item \code{717} insurance
\item \code{718} real estate, incl. real estate-insurance-law offices
\item \code{727} advertising
\item \code{728} services to dwellings and other buildings
\item \code{729} commercial research, development, and testing labs
\item \code{737} employment and temporary help agencies
\item \code{738} business management and consulting services
\item \code{739} computer programming services
\item \code{747} detective and protective services
\item \code{748} business services, n.e.c.
\item \code{749} automobile services, except repair
\item \code{757} automobile repair and related services
\item \code{758} electrical repair shops
\item \code{759} miscellaneous repair services
\item \code{769} private households
\item \code{777} hotels and motels
\item \code{778} lodging places, except hotels and motels
\item \code{779} laundering, cleaning, and other garment services
\item \code{787} beauty shops
\item \code{788} barber shops
\item \code{789} shoe repair shops
\item \code{797} dressmaking shops
\item \code{798} miscellaneous personal services
\item \code{807} theaters and motion pictures
\item \code{808} bowling alleys, billiard and pool parlors
\item \code{809} miscellaneous entertainment and recreation services
\item \code{828} offices of physicians
\item \code{829} offices of dentists
\item \code{837} offices of chiropractors
\item \code{838} hospitals
\item \code{839} convalescent institutions
\item \code{847} offices of health practitioners,n.e.c.
\item \code{848} health services, n.e.c.
\item \code{849} legal services
\item \code{857} elementary and secondary schools
\item \code{858} colleges and universities
\item \code{859} libraries
\item \code{867} educational services, n.e.c.
\item \code{868} not specified educational services
\item \code{869} museums, art galleries, and zoos
\item \code{877} religious organizations
\item \code{878} welfare services
\item \code{879} residential welfare facilities
\item \code{887} nonprofit membership organizations
\item \code{888} engineering and architectural services
\item \code{889} accounting, auditing, and bookkeeping services
\item \code{897} miscellaneous professional and related services
\item \code{907} postal service
\item \code{917} federal public administration
\item \code{927} state public administration
\item \code{937} local public administration
}
}

\keyword{variable}
