% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_601.R
\name{inthisp}
\alias{inthisp}
\title{inthisp}
\source{
General Social Survey https://gss.norc.org
}
\description{
Is interviewer spanish, hispanic or latino
}
\details{
Question 1655. Interviewer, are you Spanish, Hispanic, or Latino/Latina? Which group are you from: Mexican, Mexican American, Chicano/Chicana, Puerto Rican, Cuban, Other. Please specific other.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllll}{
   year \tab iap \tab costa rican \tab cuban \tab dominican \tab ecuadorian \tab guatemalan \tab hispanic \tab honduran \tab mexican, mexican american, chicano/a \tab no answer \tab not hispanic \tab peruvian \tab puerto rican \tab salvadorian \tab south american \tab spanish \tab venezuelan \tab argentinian \tab central american \tab colombian \tab columbian \tab latino/a \tab nicaraguan \tab other, not specified \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 9 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 65 \tab 13 \tab 2702 \tab 1 \tab 4 \tab 3 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab 2 \tab 27 \tab 24 \tab 2 \tab - \tab 1 \tab 256 \tab 19 \tab 4041 \tab 1 \tab 14 \tab 1 \tab 14 \tab 62 \tab - \tab 36 \tab 1 \tab 1 \tab 7 \tab 1 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 6 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 110 \tab - \tab 1843 \tab 9 \tab 1 \tab 1 \tab 4 \tab 22 \tab - \tab - \tab 1 \tab - \tab 18 \tab - \tab 5 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 8 \tab 13 \tab 4 \tab - \tab - \tab 1 \tab 3 \tab 83 \tab 34 \tab 1823 \tab - \tab 3 \tab 7 \tab - \tab 25 \tab 20 \tab 4 \tab - \tab - \tab - \tab - \tab 16 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab 12 \tab - \tab - \tab - \tab 19 \tab 1 \tab 103 \tab 33 \tab 1773 \tab - \tab 9 \tab - \tab - \tab 17 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab 31 \tab - \tab - \tab - \tab - \tab - \tab 91 \tab 13 \tab 2295 \tab - \tab 37 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 71 \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab 8 \tab 11 \tab 6 \tab 1 \tab - \tab 248 \tab 15 \tab 2479 \tab - \tab 91 \tab - \tab - \tab 3 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab 32 \tab - \tab - \tab - \tab - \tab - \tab 124 \tab 38 \tab 2140 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 11 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 47242 \tab 23 \tab 95 \tab 40 \tab 37 \tab 10 \tab 22 \tab 6 \tab 1080 \tab 165 \tab 19096 \tab 11 \tab 162 \tab 12 \tab 21 \tab 131 \tab 27 \tab 45 \tab 2 \tab 1 \tab 25 \tab 1 \tab 21 \tab 83 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} not hispanic
\item \code{2} mexican, mexican american, chicano/a
\item \code{3} puerto rican
\item \code{4} cuban
\item \code{5} salvadorian
\item \code{6} guatemalan
\item \code{7} panamanian
\item \code{8} nicaraguan
\item \code{9} costa rican
\item \code{10} central american
\item \code{11} honduran
\item \code{15} dominican
\item \code{16} west indian
\item \code{20} peruvian
\item \code{21} ecuadorian
\item \code{22} columbian
\item \code{23} venezuelan
\item \code{24} argentinian
\item \code{27} colombian
\item \code{30} spanish
\item \code{31} basque
\item \code{35} filipino/a
\item \code{40} latin american
\item \code{41} south american
\item \code{45} latin
\item \code{46} latino/a
\item \code{47} hispanic
\item \code{50} other, not specified
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
