% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_128.R
\name{intlblks}
\alias{intlblks}
\title{intlblks}
\source{
General Social Survey https://gss.norc.org
}
\description{
Unintelligent - intelligent
}
\details{
Question 403d. Do people in these groups tend to be unintelligent or tend to be intelligent? 3. Blacks
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1-unintelligent \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7-intelligent \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 27 \tab 93 \tab 248 \tab 600 \tab 195 \tab 77 \tab 40 \tab 61 \tab 31 \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1935 \tab 3 \tab 36 \tab 119 \tab 485 \tab 185 \tab 59 \tab 22 \tab 52 \tab 8 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 13 \tab 71 \tab 207 \tab 929 \tab 344 \tab 148 \tab 57 \tab 90 \tab 12 \tab - \tab 2832 \cr
   2000 \tab 486 \tab 30 \tab 116 \tab 334 \tab 1065 \tab 391 \tab 159 \tab 102 \tab 109 \tab 25 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 14 \tab 29 \tab 129 \tab 425 \tab 175 \tab 66 \tab 36 \tab 25 \tab 9 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 10 \tab 23 \tab 83 \tab 504 \tab 165 \tab 60 \tab 37 \tab 17 \tab 7 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 14 \tab 48 \tab 235 \tab 1049 \tab 368 \tab 140 \tab 73 \tab 53 \tab 12 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 13 \tab 27 \tab 147 \tab 699 \tab 245 \tab 94 \tab 47 \tab 33 \tab 24 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 10 \tab 37 \tab 143 \tab 775 \tab 262 \tab 89 \tab 58 \tab 41 \tab 15 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 1 \tab 36 \tab 127 \tab 712 \tab 239 \tab 96 \tab 43 \tab 29 \tab 19 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 13 \tab 39 \tab 171 \tab 915 \tab 318 \tab 114 \tab 69 \tab 21 \tab 15 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 22 \tab 39 \tab 176 \tab 1015 \tab 356 \tab 149 \tab 88 \tab 27 \tab 16 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 10 \tab 38 \tab 117 \tab 830 \tab 331 \tab 110 \tab 75 \tab 28 \tab 20 \tab - \tab 2348 \cr
   2021 \tab 1316 \tab 20 \tab 35 \tab 195 \tab 1517 \tab 509 \tab 195 \tab 200 \tab 7 \tab 7 \tab 31 \tab 4032 \cr
   2022 \tab 1168 \tab 16 \tab 29 \tab 162 \tab 1281 \tab 459 \tab 156 \tab 202 \tab 37 \tab 22 \tab 12 \tab 3544 \cr
   Total \tab 47766 \tab 216 \tab 696 \tab 2593 \tab 12801 \tab 4542 \tab 1712 \tab 1149 \tab 630 \tab 242 \tab 43 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} 1-unintelligent
\item \code{7} 7-intelligent
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
