% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_595.R
\name{isco68}
\alias{isco68}
\title{isco68}
\source{
General Social Survey https://gss.norc.org
}
\description{
Respondent's occupation, 1970 census
}
\details{
Question 1635. Respondent's occupation (International Standard Classification Codes-1968)
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 1100 \tab 1210 \tab 1310 \tab 1320 \tab 1330 \tab 1340 \tab 1391 \tab 1393 \tab 1410 \tab 1510 \tab 1591 \tab 1593 \tab 1610 \tab 1621 \tab 1730 \tab 1791 \tab 1910 \tab 1921 \tab 1930 \tab 1931 \tab 1941 \tab 1990 \tab 210 \tab 2114 \tab 2190 \tab 220 \tab 230 \tab 240 \tab 250 \tab 280 \tab 290 \tab 3000 \tab 3100 \tab 320 \tab 3210 \tab 3211 \tab 3220 \tab 3310 \tab 3311 \tab 3313 \tab 3314 \tab 3315 \tab 3390 \tab 340 \tab 3410 \tab 3420 \tab 3520 \tab 360 \tab 3600 \tab 3700 \tab 3701 \tab 3800 \tab 3801 \tab 390 \tab 3910 \tab 3911 \tab 3940 \tab 3950 \tab 3951 \tab 3990 \tab 3992 \tab 3993 \tab 4002 \tab 410 \tab 4210 \tab 4220 \tab 4221 \tab 4410 \tab 4411 \tab 4412 \tab 4432 \tab 4510 \tab 4521 \tab 5000 \tab 5201 \tab 530 \tab 5310 \tab 5320 \tab 5322 \tab 540 \tab 5400 \tab 5401 \tab 5402 \tab 5510 \tab 5520 \tab 5600 \tab 5700 \tab 5810 \tab 5820 \tab 5890 \tab 5895 \tab 5900 \tab 5991 \tab 5993 \tab 5996 \tab 610 \tab 6110 \tab 6117 \tab 620 \tab 6210 \tab 630 \tab 640 \tab 650 \tab 7000 \tab 710 \tab 7130 \tab 720 \tab 7210 \tab 7280 \tab 7320 \tab 7321 \tab 7410 \tab 7520 \tab 7540 \tab 7590 \tab 760 \tab 770 \tab 7731 \tab 7910 \tab 7911 \tab 7940 \tab 7950 \tab 7960 \tab 7990 \tab 8020 \tab 8110 \tab 830 \tab 8320 \tab 8330 \tab 8340 \tab 840 \tab 8400 \tab 8410 \tab 8422 \tab 8430 \tab 8440 \tab 8490 \tab 8493 \tab 8494 \tab 8510 \tab 8540 \tab 8550 \tab 8560 \tab 8590 \tab 8620 \tab 8710 \tab 8720 \tab 8730 \tab 8740 \tab 900 \tab 9211 \tab 9220 \tab 9240 \tab 9260 \tab 9310 \tab 9311 \tab 9390 \tab 9490 \tab 9510 \tab 9520 \tab 9530 \tab 9540 \tab 9594 \tab 9595 \tab 9690 \tab 9710 \tab 9711 \tab 9714 \tab 9730 \tab 9740 \tab 9790 \tab 9810 \tab 9830 \tab 9831 \tab 9832 \tab 9840 \tab 9851 \tab 9852 \tab 9853 \tab 9890 \tab 9971 \tab 9990 \tab 9996 \tab iap \tab 110 \tab 131 \tab 1350 \tab 1490 \tab 1710 \tab 1720 \tab 1800 \tab 270 \tab 310 \tab 3102 \tab 3391 \tab 3991 \tab 420 \tab 4222 \tab 4420 \tab 4523 \tab 5103 \tab 5321 \tab 5892 \tab 5992 \tab 6000 \tab 6001 \tab 6410 \tab 7110 \tab 750 \tab 7550 \tab 7760 \tab 790 \tab 7930 \tab 8120 \tab 8310 \tab 8351 \tab 8412 \tab 8492 \tab 8570 \tab 8800 \tab 8910 \tab 9250 \tab 9410 \tab 9850 \tab 9891 \tab 1220 \tab 1622 \tab 1630 \tab 2197 \tab 3590 \tab 4514 \tab 5920 \tab 7740 \tab 7790 \tab 820 \tab 8732 \tab 9270 \tab 9950 \tab 140 \tab 670 \tab 690 \tab 7710 \tab 9560 \tab 132 \tab 1790 \tab 1920 \tab 1924 \tab 350 \tab 510 \tab 7111 \tab 7230 \tab 810 \tab 9550 \tab 9610 \tab 8311 \tab 4430 \tab 4431 \tab 8010 \tab 260 \tab 7530 \tab 7560 \tab 7920 \tab 9542 \tab 4310 \tab 791 \tab 9590 \tab 3601 \tab 5822 \tab 9570 \tab 8200 \tab 6290 \tab 9811 \tab 120 \tab 7260 \tab Total \cr
   1972 \tab 9 \tab 1 \tab 9 \tab 31 \tab 29 \tab 7 \tab 4 \tab 7 \tab 5 \tab 1 \tab 7 \tab 4 \tab 3 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 11 \tab 1 \tab 3 \tab 1 \tab 1 \tab 4 \tab 78 \tab 3 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 9 \tab 11 \tab 3 \tab 23 \tab 57 \tab 5 \tab 23 \tab 16 \tab 5 \tab 4 \tab 7 \tab 5 \tab 2 \tab 5 \tab 3 \tab 2 \tab 1 \tab 3 \tab 11 \tab 2 \tab 12 \tab 2 \tab 7 \tab 1 \tab 8 \tab 3 \tab 1 \tab 9 \tab 52 \tab 3 \tab 1 \tab 1 \tab 2 \tab 8 \tab 6 \tab 5 \tab 7 \tab 6 \tab 2 \tab 2 \tab 56 \tab 4 \tab 12 \tab 8 \tab 3 \tab 3 \tab 33 \tab 4 \tab 4 \tab 36 \tab 6 \tab 4 \tab 18 \tab 32 \tab 6 \tab 10 \tab 4 \tab 6 \tab 4 \tab 17 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 39 \tab 1 \tab 6 \tab 13 \tab 3 \tab 4 \tab 4 \tab 26 \tab 32 \tab 1 \tab 17 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 8 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 5 \tab 7 \tab 17 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab 45 \tab 5 \tab 7 \tab 10 \tab 1 \tab 7 \tab 2 \tab 15 \tab 15 \tab 7 \tab 2 \tab 1 \tab 9 \tab 2 \tab 2 \tab 2 \tab 10 \tab 11 \tab 8 \tab 3 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 6 \tab 1 \tab 3 \tab 18 \tab 4 \tab 1 \tab 3 \tab 13 \tab 1 \tab 13 \tab 5 \tab 1 \tab 17 \tab 14 \tab 5 \tab 8 \tab 12 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 36 \tab 7 \tab 2 \tab 1 \tab 27 \tab 3 \tab 166 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 12 \tab 5 \tab 12 \tab 29 \tab 20 \tab 2 \tab 4 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 11 \tab - \tab - \tab - \tab 1 \tab 7 \tab 84 \tab 4 \tab 2 \tab 5 \tab 1 \tab 2 \tab 3 \tab 8 \tab 8 \tab 2 \tab 21 \tab 63 \tab 6 \tab 30 \tab 20 \tab 4 \tab 4 \tab 8 \tab 8 \tab 7 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 8 \tab - \tab 6 \tab 1 \tab 4 \tab 6 \tab 10 \tab 7 \tab 3 \tab 3 \tab 43 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 13 \tab 4 \tab - \tab 1 \tab 60 \tab 4 \tab 6 \tab 7 \tab 2 \tab - \tab 36 \tab 2 \tab 3 \tab 18 \tab 11 \tab 5 \tab 22 \tab 24 \tab 5 \tab 13 \tab 7 \tab 5 \tab 10 \tab 11 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 24 \tab 1 \tab 4 \tab 7 \tab 2 \tab 4 \tab - \tab 22 \tab 21 \tab 2 \tab 26 \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab 4 \tab 6 \tab 17 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 61 \tab 4 \tab 4 \tab 13 \tab - \tab 7 \tab 1 \tab 9 \tab 16 \tab 5 \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab - \tab 9 \tab 7 \tab 4 \tab 3 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 14 \tab 2 \tab - \tab 1 \tab 20 \tab - \tab 14 \tab 4 \tab - \tab 7 \tab 14 \tab 2 \tab 4 \tab 5 \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab 3 \tab 19 \tab 4 \tab - \tab 1 \tab 33 \tab 5 \tab 172 \tab 2 \tab 1 \tab 1 \tab 2 \tab 5 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 5 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 4 \tab 3 \tab 12 \tab 26 \tab 44 \tab 7 \tab 10 \tab 5 \tab 5 \tab 1 \tab 6 \tab 2 \tab 5 \tab 3 \tab - \tab 4 \tab 5 \tab 2 \tab 13 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 62 \tab - \tab 6 \tab 2 \tab 1 \tab 1 \tab 1 \tab 11 \tab 4 \tab 5 \tab 20 \tab 66 \tab 3 \tab 29 \tab 18 \tab 4 \tab 7 \tab 9 \tab 10 \tab 3 \tab 8 \tab 5 \tab 1 \tab 1 \tab 1 \tab 6 \tab 1 \tab 11 \tab - \tab 1 \tab 7 \tab 9 \tab 5 \tab 2 \tab 3 \tab 34 \tab - \tab 1 \tab 2 \tab 1 \tab 11 \tab 1 \tab 5 \tab 11 \tab 3 \tab 1 \tab 2 \tab 46 \tab 5 \tab 7 \tab 4 \tab 4 \tab - \tab 27 \tab 5 \tab 4 \tab 22 \tab 6 \tab 5 \tab 16 \tab 29 \tab 5 \tab 7 \tab 1 \tab 7 \tab 7 \tab 18 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 20 \tab 1 \tab 4 \tab 7 \tab - \tab 5 \tab - \tab 21 \tab 30 \tab 1 \tab 28 \tab 2 \tab 3 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab 2 \tab 6 \tab 7 \tab 17 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 56 \tab - \tab - \tab 2 \tab - \tab 11 \tab 3 \tab 13 \tab 20 \tab 1 \tab 1 \tab 4 \tab 5 \tab 3 \tab 2 \tab - \tab 12 \tab 4 \tab 2 \tab 3 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 20 \tab 4 \tab 1 \tab 2 \tab 21 \tab 1 \tab 11 \tab 1 \tab - \tab 15 \tab 7 \tab 5 \tab 8 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 8 \tab 13 \tab 7 \tab - \tab - \tab 33 \tab 1 \tab 133 \tab 2 \tab - \tab 3 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 6 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 6 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 11 \tab 3 \tab 4 \tab 17 \tab 26 \tab 3 \tab 4 \tab 7 \tab 2 \tab 1 \tab 2 \tab 2 \tab 3 \tab 4 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 3 \tab 3 \tab 1 \tab 9 \tab 76 \tab 1 \tab 7 \tab 3 \tab 1 \tab 1 \tab 3 \tab 12 \tab 6 \tab 1 \tab 21 \tab 66 \tab 3 \tab 31 \tab 19 \tab 5 \tab 6 \tab 6 \tab 8 \tab 2 \tab 11 \tab - \tab 3 \tab 1 \tab 2 \tab 9 \tab 2 \tab 7 \tab 1 \tab 5 \tab 10 \tab 10 \tab 11 \tab - \tab 4 \tab 27 \tab 2 \tab 1 \tab 2 \tab - \tab 11 \tab - \tab 5 \tab 8 \tab 4 \tab 2 \tab - \tab 54 \tab 3 \tab 6 \tab 3 \tab 1 \tab - \tab 35 \tab 4 \tab 7 \tab 25 \tab 9 \tab 8 \tab 18 \tab 23 \tab 1 \tab 12 \tab 1 \tab 7 \tab 7 \tab 14 \tab - \tab 3 \tab 1 \tab - \tab - \tab 22 \tab 1 \tab 2 \tab 7 \tab - \tab 2 \tab - \tab 20 \tab 31 \tab 1 \tab 32 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 5 \tab 2 \tab 2 \tab 3 \tab 2 \tab 8 \tab 1 \tab 7 \tab 7 \tab 38 \tab - \tab 4 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 40 \tab 6 \tab 4 \tab 11 \tab - \tab 16 \tab 1 \tab 13 \tab 20 \tab 3 \tab - \tab 3 \tab 8 \tab 1 \tab 1 \tab - \tab 6 \tab 7 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 8 \tab - \tab 1 \tab 10 \tab 6 \tab 2 \tab 1 \tab 22 \tab - \tab 11 \tab 3 \tab - \tab 8 \tab 12 \tab 4 \tab 8 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 5 \tab 20 \tab 9 \tab - \tab 1 \tab 25 \tab 1 \tab 141 \tab 4 \tab - \tab 2 \tab 2 \tab 5 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 8 \tab 2 \tab - \tab - \tab - \tab - \tab 7 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 9 \tab 4 \tab 7 \tab 30 \tab 29 \tab - \tab 3 \tab 7 \tab 7 \tab - \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 8 \tab 3 \tab 1 \tab 1 \tab 1 \tab 8 \tab 68 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 9 \tab 4 \tab 4 \tab 26 \tab 60 \tab 8 \tab 37 \tab 23 \tab 1 \tab 6 \tab 3 \tab 8 \tab 5 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 14 \tab 2 \tab 4 \tab 3 \tab 15 \tab 9 \tab 1 \tab 5 \tab 31 \tab - \tab - \tab 3 \tab 1 \tab 11 \tab 1 \tab 3 \tab 12 \tab 6 \tab - \tab - \tab 63 \tab 9 \tab 7 \tab 8 \tab 1 \tab 3 \tab 32 \tab 3 \tab 5 \tab 22 \tab 7 \tab 3 \tab 13 \tab 29 \tab 9 \tab 8 \tab 6 \tab 6 \tab 6 \tab 11 \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 37 \tab - \tab 4 \tab 11 \tab - \tab 1 \tab - \tab 22 \tab 26 \tab - \tab 17 \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab 2 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 2 \tab 5 \tab 30 \tab 1 \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab 49 \tab 5 \tab 9 \tab 4 \tab - \tab 14 \tab 1 \tab 15 \tab 25 \tab 5 \tab - \tab 2 \tab 3 \tab 2 \tab - \tab - \tab 5 \tab 13 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 6 \tab - \tab 3 \tab 16 \tab 4 \tab 1 \tab 1 \tab 7 \tab - \tab 9 \tab 4 \tab 3 \tab 13 \tab 11 \tab 3 \tab 10 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 17 \tab 8 \tab - \tab - \tab 29 \tab 2 \tab 147 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 8 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 11 \tab 1 \tab 10 \tab 31 \tab 35 \tab 3 \tab 5 \tab 7 \tab 9 \tab - \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 9 \tab 1 \tab 1 \tab - \tab 2 \tab 5 \tab 70 \tab 3 \tab 2 \tab 2 \tab 3 \tab 6 \tab 1 \tab 9 \tab 7 \tab 3 \tab 16 \tab 65 \tab 5 \tab 23 \tab 26 \tab 7 \tab 4 \tab 5 \tab 9 \tab 3 \tab 6 \tab 3 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab 10 \tab - \tab 4 \tab 6 \tab 10 \tab 4 \tab 2 \tab 1 \tab 33 \tab 1 \tab 2 \tab - \tab 3 \tab 5 \tab - \tab 5 \tab 5 \tab 10 \tab 1 \tab 3 \tab 58 \tab 1 \tab 10 \tab 5 \tab - \tab - \tab 41 \tab 7 \tab 6 \tab 26 \tab 7 \tab 3 \tab 16 \tab 21 \tab 3 \tab 14 \tab - \tab 1 \tab 8 \tab 19 \tab 4 \tab - \tab - \tab - \tab 2 \tab 23 \tab 1 \tab 2 \tab 10 \tab 3 \tab - \tab - \tab 13 \tab 26 \tab 3 \tab 21 \tab 1 \tab - \tab - \tab - \tab 4 \tab 2 \tab - \tab 1 \tab 3 \tab 3 \tab 4 \tab 1 \tab 5 \tab 8 \tab 25 \tab 3 \tab 5 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 61 \tab 4 \tab 7 \tab 4 \tab - \tab 19 \tab - \tab 15 \tab 27 \tab 3 \tab - \tab 2 \tab 7 \tab 2 \tab 1 \tab - \tab 3 \tab 14 \tab 4 \tab 2 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 13 \tab 1 \tab 4 \tab 18 \tab 4 \tab 1 \tab - \tab 22 \tab - \tab 21 \tab 1 \tab - \tab 15 \tab 22 \tab 1 \tab 13 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 18 \tab 15 \tab - \tab - \tab 26 \tab 3 \tab 114 \tab 3 \tab 1 \tab 2 \tab - \tab 4 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 7 \tab - \tab 4 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 8 \tab 3 \tab 6 \tab 20 \tab 49 \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 4 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 14 \tab 2 \tab 1 \tab - \tab 4 \tab 11 \tab 65 \tab 3 \tab 2 \tab 4 \tab 2 \tab - \tab 1 \tab 13 \tab 8 \tab 3 \tab 21 \tab 71 \tab 3 \tab 43 \tab 16 \tab 4 \tab 2 \tab 5 \tab 12 \tab 3 \tab 3 \tab 2 \tab - \tab 4 \tab 1 \tab 2 \tab 2 \tab 10 \tab 2 \tab 6 \tab 5 \tab 9 \tab 4 \tab 1 \tab 4 \tab 20 \tab 1 \tab 1 \tab - \tab 1 \tab 11 \tab 2 \tab 3 \tab 10 \tab 9 \tab 1 \tab 2 \tab 63 \tab 5 \tab 12 \tab 3 \tab 3 \tab - \tab 39 \tab 10 \tab 10 \tab 22 \tab 6 \tab 5 \tab 12 \tab 26 \tab 3 \tab 17 \tab 6 \tab 6 \tab 5 \tab 21 \tab - \tab - \tab - \tab - \tab 4 \tab 15 \tab - \tab 7 \tab 10 \tab 2 \tab 4 \tab - \tab 22 \tab 28 \tab - \tab 22 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 5 \tab - \tab 3 \tab 4 \tab - \tab 1 \tab 2 \tab 4 \tab 2 \tab 31 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 6 \tab 3 \tab 39 \tab 5 \tab 7 \tab 5 \tab 1 \tab 10 \tab - \tab 16 \tab 19 \tab 3 \tab 4 \tab 2 \tab 12 \tab 7 \tab 3 \tab - \tab 4 \tab 8 \tab 3 \tab - \tab 1 \tab 1 \tab 5 \tab 1 \tab 2 \tab 4 \tab - \tab 2 \tab 23 \tab - \tab 1 \tab 2 \tab 17 \tab 2 \tab 8 \tab 4 \tab 1 \tab 14 \tab 14 \tab 5 \tab 6 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 8 \tab 20 \tab 11 \tab - \tab - \tab 30 \tab 2 \tab 128 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 13 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 15 \tab 8 \tab 7 \tab 23 \tab 28 \tab 3 \tab 5 \tab 2 \tab - \tab - \tab 5 \tab 3 \tab 3 \tab 3 \tab 1 \tab - \tab 3 \tab - \tab 13 \tab - \tab 1 \tab 3 \tab 1 \tab 5 \tab 77 \tab 1 \tab 3 \tab - \tab 2 \tab 5 \tab 3 \tab 16 \tab 8 \tab 6 \tab 17 \tab 56 \tab 5 \tab 25 \tab 10 \tab 5 \tab 5 \tab 5 \tab 14 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab 5 \tab 1 \tab 7 \tab 6 \tab 7 \tab 9 \tab - \tab 2 \tab 45 \tab - \tab - \tab 2 \tab 1 \tab 10 \tab 7 \tab 3 \tab 10 \tab 10 \tab 4 \tab 2 \tab 44 \tab 8 \tab 13 \tab 11 \tab 4 \tab - \tab 34 \tab 1 \tab 5 \tab 15 \tab 10 \tab 4 \tab 21 \tab 30 \tab 3 \tab 9 \tab 4 \tab 5 \tab 3 \tab 14 \tab - \tab - \tab - \tab - \tab 3 \tab 24 \tab - \tab 9 \tab 8 \tab - \tab 2 \tab 1 \tab 20 \tab 27 \tab 1 \tab 27 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 7 \tab 3 \tab 1 \tab 5 \tab 1 \tab 2 \tab 5 \tab 23 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 4 \tab - \tab 37 \tab 13 \tab 13 \tab 13 \tab 1 \tab 9 \tab 3 \tab 13 \tab 17 \tab 2 \tab 1 \tab 1 \tab 10 \tab 1 \tab 2 \tab - \tab 2 \tab 12 \tab 2 \tab - \tab 3 \tab 1 \tab 4 \tab - \tab - \tab 5 \tab - \tab 3 \tab 11 \tab 3 \tab 1 \tab 1 \tab 12 \tab - \tab 8 \tab 3 \tab - \tab 5 \tab 9 \tab 6 \tab 5 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 16 \tab 9 \tab 1 \tab - \tab 31 \tab 1 \tab 119 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab 9 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 15 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 5 \tab 2 \tab - \tab - \tab 5 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 5 \tab 10 \tab 7 \tab 31 \tab 46 \tab 4 \tab 4 \tab 6 \tab 4 \tab - \tab - \tab 2 \tab 6 \tab 2 \tab - \tab 5 \tab 2 \tab - \tab 19 \tab 1 \tab 2 \tab - \tab 3 \tab 7 \tab 64 \tab - \tab - \tab 2 \tab - \tab 4 \tab 1 \tab 8 \tab 8 \tab 3 \tab 14 \tab 60 \tab 2 \tab 34 \tab 23 \tab 7 \tab 8 \tab 7 \tab 7 \tab 8 \tab 3 \tab 13 \tab - \tab 1 \tab - \tab 7 \tab 1 \tab 10 \tab 1 \tab 4 \tab 7 \tab 19 \tab 2 \tab - \tab 4 \tab 80 \tab 1 \tab 2 \tab - \tab - \tab 23 \tab 1 \tab - \tab 9 \tab 4 \tab 1 \tab 4 \tab 59 \tab 2 \tab 8 \tab 1 \tab - \tab 3 \tab 51 \tab 3 \tab 10 \tab 54 \tab 21 \tab 7 \tab 13 \tab 55 \tab 7 \tab 22 \tab 2 \tab 5 \tab 12 \tab 11 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 21 \tab - \tab 8 \tab 12 \tab 1 \tab 4 \tab - \tab 26 \tab 37 \tab - \tab 46 \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 3 \tab 10 \tab - \tab 1 \tab 10 \tab 22 \tab 3 \tab 4 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 76 \tab 6 \tab 11 \tab 1 \tab - \tab 17 \tab 3 \tab 26 \tab 22 \tab 7 \tab - \tab 1 \tab 13 \tab 1 \tab 5 \tab - \tab 6 \tab 9 \tab 6 \tab 1 \tab 2 \tab 2 \tab 6 \tab - \tab 3 \tab 7 \tab 1 \tab 4 \tab 24 \tab 5 \tab 1 \tab - \tab 14 \tab 3 \tab 20 \tab 1 \tab - \tab 16 \tab 19 \tab 3 \tab 9 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 24 \tab 7 \tab - \tab - \tab 48 \tab 3 \tab 137 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab 1 \tab - \tab 15 \tab - \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 6 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 5 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 18 \tab 4 \tab 9 \tab 27 \tab 33 \tab 2 \tab 2 \tab 7 \tab 5 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab 5 \tab - \tab 14 \tab - \tab 2 \tab 3 \tab 1 \tab 9 \tab 100 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 21 \tab 15 \tab 2 \tab 17 \tab 85 \tab 6 \tab 39 \tab 29 \tab 3 \tab 2 \tab 6 \tab 16 \tab 4 \tab 2 \tab 11 \tab 2 \tab 1 \tab - \tab 2 \tab 5 \tab 4 \tab 2 \tab 10 \tab 12 \tab 14 \tab 3 \tab 1 \tab 3 \tab 33 \tab 1 \tab - \tab 1 \tab 2 \tab 14 \tab 1 \tab 5 \tab 8 \tab 6 \tab 3 \tab 3 \tab 58 \tab - \tab 4 \tab 21 \tab - \tab 1 \tab 37 \tab 7 \tab 14 \tab 11 \tab 11 \tab 1 \tab 15 \tab 32 \tab 6 \tab 10 \tab 1 \tab 4 \tab 8 \tab 12 \tab - \tab 2 \tab - \tab - \tab 2 \tab 33 \tab - \tab 14 \tab 7 \tab 4 \tab 6 \tab 2 \tab 12 \tab 29 \tab - \tab 29 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 4 \tab 2 \tab 5 \tab 2 \tab 1 \tab 6 \tab 23 \tab 2 \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 34 \tab 8 \tab 9 \tab 8 \tab - \tab 11 \tab 3 \tab 10 \tab 25 \tab 6 \tab - \tab 4 \tab 12 \tab 1 \tab 2 \tab - \tab 7 \tab 10 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 6 \tab - \tab 1 \tab 12 \tab 5 \tab 2 \tab 2 \tab 14 \tab - \tab 10 \tab 1 \tab 1 \tab 11 \tab 10 \tab 3 \tab 6 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab 3 \tab 17 \tab 7 \tab - \tab 1 \tab 22 \tab 2 \tab 106 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 11 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 4 \tab 5 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 16 \tab 5 \tab 6 \tab 29 \tab 25 \tab 5 \tab 1 \tab 9 \tab 2 \tab 1 \tab 2 \tab 6 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 11 \tab 1 \tab 3 \tab 3 \tab 1 \tab 6 \tab 92 \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab 4 \tab 16 \tab 14 \tab 1 \tab 13 \tab 50 \tab 2 \tab 24 \tab 26 \tab 6 \tab 5 \tab 30 \tab 15 \tab 5 \tab 3 \tab 11 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 7 \tab - \tab 4 \tab 13 \tab 14 \tab 8 \tab - \tab 11 \tab 22 \tab 1 \tab - \tab 4 \tab 1 \tab 9 \tab 7 \tab 3 \tab 2 \tab 6 \tab 3 \tab 3 \tab 56 \tab - \tab 13 \tab 15 \tab - \tab 1 \tab 26 \tab 11 \tab 7 \tab 6 \tab 14 \tab 8 \tab 13 \tab 39 \tab 5 \tab 14 \tab 2 \tab 6 \tab 4 \tab 19 \tab - \tab 2 \tab - \tab - \tab 5 \tab 21 \tab - \tab 5 \tab 9 \tab - \tab 4 \tab - \tab 13 \tab 34 \tab - \tab 20 \tab - \tab - \tab - \tab - \tab 1 \tab 6 \tab - \tab 1 \tab 4 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 16 \tab - \tab 6 \tab 2 \tab 3 \tab 3 \tab 1 \tab - \tab 36 \tab 10 \tab 13 \tab 8 \tab - \tab 11 \tab 2 \tab 9 \tab 16 \tab - \tab - \tab 1 \tab 5 \tab 2 \tab - \tab - \tab 6 \tab 10 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab 21 \tab 1 \tab - \tab 1 \tab 17 \tab - \tab 13 \tab 1 \tab - \tab 10 \tab 11 \tab 1 \tab 3 \tab 7 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 15 \tab 5 \tab - \tab 1 \tab 56 \tab - \tab 94 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 6 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 14 \tab 8 \tab 12 \tab 17 \tab 28 \tab 2 \tab 4 \tab 7 \tab 2 \tab - \tab 3 \tab 2 \tab 4 \tab 4 \tab - \tab 2 \tab 2 \tab 4 \tab 9 \tab 2 \tab 1 \tab 4 \tab 2 \tab 12 \tab 105 \tab 6 \tab 7 \tab 2 \tab 7 \tab 5 \tab 1 \tab 16 \tab 12 \tab 2 \tab 15 \tab 59 \tab 5 \tab 35 \tab 18 \tab 7 \tab 4 \tab 5 \tab 6 \tab 1 \tab 2 \tab 17 \tab 1 \tab 1 \tab 3 \tab 7 \tab 2 \tab 6 \tab 2 \tab 3 \tab 6 \tab 6 \tab 12 \tab 3 \tab 4 \tab 46 \tab 1 \tab - \tab 2 \tab 3 \tab 18 \tab 3 \tab 3 \tab 10 \tab 9 \tab 2 \tab 4 \tab 46 \tab 1 \tab 11 \tab 6 \tab 2 \tab - \tab 45 \tab 7 \tab 9 \tab 14 \tab 12 \tab 2 \tab 13 \tab 44 \tab 5 \tab 15 \tab 3 \tab 8 \tab 6 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab 6 \tab 24 \tab - \tab 5 \tab 12 \tab 6 \tab 2 \tab - \tab 23 \tab 28 \tab 1 \tab 25 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 5 \tab 18 \tab 1 \tab 3 \tab - \tab - \tab 3 \tab 4 \tab 1 \tab 30 \tab 9 \tab 7 \tab 3 \tab 1 \tab 11 \tab 2 \tab 25 \tab 13 \tab 2 \tab - \tab 2 \tab 19 \tab 1 \tab - \tab - \tab 3 \tab 8 \tab 3 \tab 2 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 7 \tab 6 \tab 2 \tab - \tab 10 \tab 1 \tab 14 \tab 1 \tab - \tab 9 \tab 11 \tab 3 \tab 7 \tab 5 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 24 \tab 1 \tab - \tab - \tab 35 \tab - \tab 81 \tab - \tab 4 \tab 2 \tab - \tab 7 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 12 \tab 3 \tab 15 \tab 19 \tab 25 \tab 2 \tab 5 \tab 6 \tab 1 \tab - \tab 4 \tab 2 \tab 5 \tab 2 \tab - \tab 1 \tab 3 \tab 2 \tab 7 \tab 5 \tab 5 \tab 3 \tab 1 \tab 8 \tab 115 \tab 3 \tab 4 \tab 2 \tab - \tab 5 \tab 2 \tab 12 \tab 8 \tab 3 \tab 16 \tab 50 \tab 5 \tab 16 \tab 31 \tab 8 \tab 2 \tab 3 \tab 8 \tab 11 \tab 3 \tab 9 \tab - \tab - \tab - \tab 6 \tab 1 \tab 9 \tab 1 \tab 6 \tab 8 \tab 7 \tab 10 \tab 1 \tab 4 \tab 30 \tab - \tab 1 \tab - \tab 3 \tab 17 \tab 4 \tab 2 \tab 8 \tab 8 \tab 2 \tab 1 \tab 65 \tab 1 \tab 13 \tab 21 \tab 2 \tab - \tab 29 \tab 5 \tab 8 \tab 2 \tab 10 \tab 4 \tab 11 \tab 22 \tab 8 \tab 17 \tab 1 \tab 8 \tab 13 \tab 7 \tab 2 \tab 2 \tab - \tab 1 \tab 6 \tab 28 \tab - \tab 5 \tab 5 \tab 1 \tab 7 \tab - \tab 21 \tab 39 \tab - \tab 31 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 13 \tab 3 \tab 4 \tab 2 \tab 1 \tab 1 \tab 5 \tab 11 \tab 2 \tab 4 \tab - \tab - \tab 3 \tab 4 \tab - \tab 31 \tab 7 \tab 6 \tab 10 \tab - \tab 11 \tab 1 \tab 13 \tab 16 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 6 \tab 10 \tab 3 \tab 1 \tab - \tab 1 \tab 5 \tab - \tab - \tab 7 \tab - \tab - \tab 12 \tab 2 \tab 1 \tab - \tab 16 \tab - \tab 5 \tab 3 \tab - \tab 6 \tab 8 \tab 5 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab 19 \tab 3 \tab - \tab 1 \tab 37 \tab 1 \tab 102 \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab 6 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1470 \cr
   1987 \tab 33 \tab 6 \tab 11 \tab 32 \tab 26 \tab 4 \tab 3 \tab 10 \tab 2 \tab 1 \tab 3 \tab 1 \tab 4 \tab 2 \tab 1 \tab 3 \tab 2 \tab 2 \tab 15 \tab - \tab 3 \tab 1 \tab 2 \tab 8 \tab 154 \tab 3 \tab 6 \tab 3 \tab 1 \tab 1 \tab 4 \tab 8 \tab 14 \tab 2 \tab 10 \tab 63 \tab 1 \tab 14 \tab 28 \tab 5 \tab 5 \tab 3 \tab 4 \tab 9 \tab 1 \tab 18 \tab 1 \tab - \tab 1 \tab 8 \tab - \tab 7 \tab 2 \tab 8 \tab 16 \tab 7 \tab 12 \tab 2 \tab 6 \tab 54 \tab 1 \tab - \tab - \tab 1 \tab 24 \tab 2 \tab 2 \tab 8 \tab 6 \tab 1 \tab 1 \tab 59 \tab 4 \tab 10 \tab 39 \tab 2 \tab 1 \tab 38 \tab 1 \tab 11 \tab 22 \tab 16 \tab 11 \tab 22 \tab 34 \tab 6 \tab 20 \tab 3 \tab 4 \tab 14 \tab 9 \tab - \tab 4 \tab 1 \tab 1 \tab 5 \tab 31 \tab 1 \tab 3 \tab 11 \tab 1 \tab 2 \tab - \tab 20 \tab 41 \tab - \tab 62 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 7 \tab - \tab 3 \tab 3 \tab 1 \tab 3 \tab 23 \tab 2 \tab 1 \tab 3 \tab 1 \tab 5 \tab - \tab 2 \tab 45 \tab 11 \tab 13 \tab 4 \tab - \tab 10 \tab 2 \tab 12 \tab 20 \tab 5 \tab 1 \tab 1 \tab 12 \tab 4 \tab - \tab 1 \tab 5 \tab 9 \tab 3 \tab 1 \tab 4 \tab 3 \tab 4 \tab 1 \tab 2 \tab 6 \tab - \tab 1 \tab 14 \tab 4 \tab - \tab 5 \tab 21 \tab 1 \tab 25 \tab 2 \tab 1 \tab 13 \tab 10 \tab 10 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 17 \tab 11 \tab - \tab 1 \tab 43 \tab - \tab 103 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 1 \tab 4 \tab 4 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1819 \cr
   1988 \tab 19 \tab 6 \tab 10 \tab 16 \tab 27 \tab 6 \tab 4 \tab 4 \tab 5 \tab 1 \tab 4 \tab 2 \tab 5 \tab 3 \tab - \tab 2 \tab 2 \tab 2 \tab 9 \tab 3 \tab 3 \tab 3 \tab - \tab 8 \tab 109 \tab 3 \tab 5 \tab 3 \tab 1 \tab 5 \tab 6 \tab 11 \tab 6 \tab 7 \tab 11 \tab 51 \tab 7 \tab 22 \tab 29 \tab 5 \tab 5 \tab 3 \tab 14 \tab 4 \tab 3 \tab 2 \tab 1 \tab 1 \tab 3 \tab 6 \tab - \tab 9 \tab - \tab 5 \tab 8 \tab 12 \tab 13 \tab 1 \tab - \tab 36 \tab - \tab - \tab 3 \tab 3 \tab 13 \tab 2 \tab - \tab 5 \tab 6 \tab 1 \tab 6 \tab 78 \tab 3 \tab 8 \tab 14 \tab 1 \tab - \tab 25 \tab 7 \tab 9 \tab 10 \tab 19 \tab 6 \tab 6 \tab 45 \tab 4 \tab 10 \tab 4 \tab 8 \tab 8 \tab 7 \tab - \tab 5 \tab - \tab 1 \tab 6 \tab 16 \tab 1 \tab 5 \tab 5 \tab 2 \tab 1 \tab - \tab 19 \tab 34 \tab - \tab 28 \tab 3 \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 2 \tab - \tab 6 \tab 2 \tab 1 \tab 1 \tab 23 \tab 1 \tab 1 \tab - \tab 3 \tab 4 \tab 1 \tab 1 \tab 26 \tab 10 \tab 3 \tab 5 \tab 1 \tab 16 \tab 3 \tab 15 \tab 22 \tab 3 \tab - \tab 2 \tab 8 \tab 5 \tab 1 \tab 1 \tab 1 \tab 3 \tab 4 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab 6 \tab - \tab 2 \tab 10 \tab 3 \tab 1 \tab 5 \tab 19 \tab - \tab 14 \tab 4 \tab 1 \tab 6 \tab 4 \tab 3 \tab 5 \tab 4 \tab - \tab 1 \tab - \tab - \tab 1 \tab 10 \tab 19 \tab 5 \tab - \tab - \tab 27 \tab 1 \tab 87 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 6 \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 2 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 14 \tab 16 \tab 20 \tab 20 \tab 26 \tab 7 \tab 4 \tab 8 \tab 2 \tab 1 \tab 6 \tab 2 \tab 4 \tab 2 \tab 2 \tab 2 \tab 4 \tab 1 \tab 10 \tab - \tab 5 \tab - \tab 2 \tab 13 \tab 110 \tab 7 \tab 5 \tab 3 \tab 2 \tab 4 \tab 1 \tab 19 \tab 15 \tab 2 \tab 6 \tab 62 \tab 3 \tab 17 \tab 30 \tab 4 \tab 5 \tab 9 \tab 7 \tab 2 \tab 3 \tab 6 \tab - \tab 2 \tab - \tab 5 \tab 2 \tab 5 \tab - \tab 5 \tab 10 \tab 9 \tab 11 \tab 2 \tab 2 \tab 50 \tab 1 \tab - \tab 1 \tab - \tab 22 \tab 2 \tab 1 \tab 5 \tab 6 \tab 3 \tab 6 \tab 63 \tab - \tab 11 \tab 5 \tab 2 \tab - \tab 30 \tab 5 \tab 14 \tab 11 \tab 8 \tab 5 \tab 14 \tab 36 \tab 9 \tab 8 \tab 2 \tab 8 \tab 8 \tab 9 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 18 \tab 1 \tab 5 \tab 3 \tab 1 \tab 4 \tab - \tab 11 \tab 41 \tab - \tab 32 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab 5 \tab 2 \tab 7 \tab 1 \tab 7 \tab 1 \tab 9 \tab - \tab 3 \tab - \tab 2 \tab 7 \tab 2 \tab - \tab 34 \tab 12 \tab 6 \tab 6 \tab - \tab 13 \tab 5 \tab 18 \tab 13 \tab 1 \tab 1 \tab 3 \tab 14 \tab 3 \tab - \tab - \tab 5 \tab 9 \tab 7 \tab 6 \tab 5 \tab 5 \tab 1 \tab 1 \tab 2 \tab 8 \tab - \tab - \tab 8 \tab 3 \tab - \tab 1 \tab 15 \tab 2 \tab 13 \tab 1 \tab 2 \tab 8 \tab 5 \tab 2 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 5 \tab 20 \tab 8 \tab - \tab - \tab 28 \tab 1 \tab 96 \tab 2 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 8 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1537 \cr
   1990 \tab 18 \tab 8 \tab 19 \tab 16 \tab 29 \tab 1 \tab 6 \tab 4 \tab 2 \tab 2 \tab 4 \tab 5 \tab 7 \tab 2 \tab - \tab - \tab 6 \tab 4 \tab 16 \tab 2 \tab 2 \tab 2 \tab 5 \tab 9 \tab 81 \tab 2 \tab 7 \tab 4 \tab - \tab 5 \tab 2 \tab 15 \tab 5 \tab 1 \tab 10 \tab 51 \tab 4 \tab 18 \tab 21 \tab 3 \tab 5 \tab 5 \tab 6 \tab 10 \tab 5 \tab 6 \tab - \tab 1 \tab - \tab 6 \tab 2 \tab 5 \tab - \tab 7 \tab 6 \tab 13 \tab 11 \tab 2 \tab 2 \tab 47 \tab - \tab 1 \tab - \tab 2 \tab 20 \tab 1 \tab 1 \tab 8 \tab 14 \tab 1 \tab 3 \tab 52 \tab 3 \tab 19 \tab 10 \tab 3 \tab - \tab 19 \tab 3 \tab 6 \tab 6 \tab 12 \tab 2 \tab 13 \tab 29 \tab - \tab 11 \tab 2 \tab 2 \tab 8 \tab 8 \tab 2 \tab 8 \tab 1 \tab - \tab 4 \tab 16 \tab 3 \tab 5 \tab 6 \tab 1 \tab 2 \tab - \tab 14 \tab 31 \tab 1 \tab 27 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 4 \tab 1 \tab 13 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 42 \tab 8 \tab 5 \tab 6 \tab - \tab 8 \tab 4 \tab 10 \tab 15 \tab 2 \tab 2 \tab - \tab 6 \tab 3 \tab 2 \tab - \tab 5 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab 18 \tab 3 \tab - \tab - \tab 16 \tab 2 \tab 12 \tab - \tab 1 \tab 11 \tab 12 \tab - \tab 4 \tab 6 \tab - \tab - \tab - \tab 1 \tab 1 \tab 6 \tab 27 \tab 3 \tab - \tab - \tab 27 \tab - \tab 72 \tab 2 \tab - \tab 5 \tab 2 \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab 5 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 228 \tab 94 \tab 176 \tab 414 \tab 525 \tab 58 \tab 70 \tab 101 \tab 56 \tab 10 \tab 52 \tab 45 \tab 61 \tab 38 \tab 8 \tab 32 \tab 46 \tab 22 \tab 192 \tab 22 \tab 37 \tab 28 \tab 29 \tab 137 \tab 1510 \tab 49 \tab 63 \tab 43 \tab 24 \tab 50 \tab 38 \tab 213 \tab 153 \tab 50 \tab 277 \tab 1035 \tab 73 \tab 460 \tab 383 \tab 83 \tab 79 \tab 119 \tab 157 \tab 82 \tab 66 \tab 111 \tab 13 \tab 18 \tab 16 \tab 96 \tab 24 \tab 137 \tab 17 \tab 90 \tab 130 \tab 179 \tab 134 \tab 22 \tab 67 \tab 683 \tab 13 \tab 10 \tab 21 \tab 25 \tab 229 \tab 40 \tab 47 \tab 139 \tab 117 \tab 28 \tab 43 \tab 980 \tab 53 \tab 170 \tab 181 \tab 30 \tab 12 \tab 577 \tab 85 \tab 132 \tab 322 \tab 185 \tab 83 \tab 256 \tab 550 \tab 85 \tab 217 \tab 49 \tab 96 \tab 131 \tab 211 \tab 17 \tab 34 \tab 5 \tab 7 \tab 53 \tab 412 \tab 11 \tab 93 \tab 143 \tab 27 \tab 54 \tab 7 \tab 325 \tab 535 \tab 11 \tab 490 \tab 20 \tab 9 \tab 16 \tab 6 \tab 15 \tab 56 \tab 15 \tab 43 \tab 51 \tab 27 \tab 71 \tab 22 \tab 53 \tab 81 \tab 356 \tab 20 \tab 51 \tab 13 \tab 17 \tab 40 \tab 41 \tab 14 \tab 742 \tab 123 \tab 124 \tab 113 \tab 5 \tab 201 \tab 36 \tab 247 \tab 321 \tab 55 \tab 14 \tab 34 \tab 148 \tab 42 \tab 23 \tab 4 \tab 95 \tab 146 \tab 59 \tab 28 \tab 34 \tab 34 \tab 41 \tab 4 \tab 16 \tab 99 \tab 3 \tab 30 \tab 256 \tab 59 \tab 15 \tab 25 \tab 276 \tab 13 \tab 221 \tab 39 \tab 11 \tab 184 \tab 193 \tab 61 \tab 102 \tab 64 \tab 7 \tab 21 \tab 2 \tab 8 \tab 14 \tab 73 \tab 341 \tab 120 \tab 3 \tab 7 \tab 557 \tab 26 \tab 48123 \tab 33 \tab 11 \tab 27 \tab 13 \tab 41 \tab 6 \tab 19 \tab 2 \tab 13 \tab 44 \tab 20 \tab 12 \tab 6 \tab 4 \tab 24 \tab 11 \tab 8 \tab 57 \tab 20 \tab 7 \tab 12 \tab 5 \tab 9 \tab 99 \tab 7 \tab 18 \tab 38 \tab 4 \tab 5 \tab 13 \tab 6 \tab 30 \tab 26 \tab 7 \tab 46 \tab 9 \tab 8 \tab 10 \tab 4 \tab 36 \tab 16 \tab 7 \tab 27 \tab 27 \tab 8 \tab 21 \tab 14 \tab 15 \tab 15 \tab 6 \tab 2 \tab 8 \tab 18 \tab 9 \tab 10 \tab 23 \tab 16 \tab 1 \tab 5 \tab 2 \tab 8 \tab 2 \tab 8 \tab 5 \tab 16 \tab 2 \tab 3 \tab 4 \tab 6 \tab 4 \tab 2 \tab 2 \tab 8 \tab 5 \tab 5 \tab 2 \tab 4 \tab 1 \tab 9 \tab 2 \tab 7 \tab 2 \tab 1 \tab 2 \tab 6 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
