% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_595.R
\name{isco681}
\alias{isco681}
\title{isco681}
\source{
General Social Survey https://gss.norc.org
}
\description{
Respondent's occupation, 1980 census
}
\details{
Question 1638. Respondent's occupation (International Standard Classification Codes-1968)
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 110 \tab 1100 \tab 1219 \tab 1310 \tab 1320 \tab 1330 \tab 1340 \tab 1350 \tab 1390 \tab 1393 \tab 140 \tab 1410 \tab 1490 \tab 1519 \tab 1590 \tab 1593 \tab 1610 \tab 1621 \tab 1710 \tab 1720 \tab 1730 \tab 1790 \tab 1799 \tab 1910 \tab 1921 \tab 1930 \tab 1939 \tab 1940 \tab 1999 \tab 2010 \tab 210 \tab 2110 \tab 2190 \tab 2192 \tab 220 \tab 230 \tab 240 \tab 250 \tab 260 \tab 280 \tab 290 \tab 3000 \tab 310 \tab 3102 \tab 320 \tab 3211 \tab 3219 \tab 3220 \tab 3311 \tab 3313 \tab 3319 \tab 3391 \tab 3399 \tab 340 \tab 3410 \tab 3420 \tab 3520 \tab 3590 \tab 3609 \tab 3700 \tab 3709 \tab 3809 \tab 390 \tab 3911 \tab 3919 \tab 3920 \tab 3939 \tab 3949 \tab 3951 \tab 3959 \tab 3990 \tab 3999 \tab 410 \tab 421 \tab 4210 \tab 4220 \tab 4319 \tab 4320 \tab 4400 \tab 4410 \tab 4411 \tab 4412 \tab 4419 \tab 4431 \tab 4432 \tab 4519 \tab 4520 \tab 4521 \tab 4523 \tab 4900 \tab 510 \tab 5200 \tab 530 \tab 5300 \tab 5312 \tab 5319 \tab 5320 \tab 5321 \tab 5322 \tab 5329 \tab 540 \tab 5400 \tab 5401 \tab 549 \tab 5500 \tab 5510 \tab 5520 \tab 5600 \tab 5700 \tab 5702 \tab 5709 \tab 5810 \tab 5829 \tab 5890 \tab 5891 \tab 5895 \tab 5900 \tab 5990 \tab 5991 \tab 5999 \tab 6000 \tab 610 \tab 6110 \tab 6120 \tab 620 \tab 6210 \tab 6270 \tab 630 \tab 6310 \tab 6319 \tab 640 \tab 6410 \tab 670 \tab 690 \tab 7000 \tab 710 \tab 7119 \tab 7240 \tab 7270 \tab 7320 \tab 7410 \tab 7430 \tab 7520 \tab 7540 \tab 760 \tab 7730 \tab 7760 \tab 7910 \tab 7919 \tab 7950 \tab 7960 \tab 7990 \tab 8010 \tab 8110 \tab 8190 \tab 830 \tab 8329 \tab 8340 \tab 8351 \tab 8421 \tab 8422 \tab 8430 \tab 8440 \tab 849 \tab 8490 \tab 8493 \tab 8494 \tab 8499 \tab 8520 \tab 8530 \tab 8540 \tab 8559 \tab 8560 \tab 8610 \tab 8620 \tab 8719 \tab 8720 \tab 8730 \tab 8739 \tab 8800 \tab 8910 \tab 8930 \tab 8950 \tab 9210 \tab 9220 \tab 9270 \tab 9310 \tab 9390 \tab 9430 \tab 9490 \tab 9499 \tab 9510 \tab 9520 \tab 9530 \tab 9594 \tab 9599 \tab 9610 \tab 9690 \tab 9710 \tab 9712 \tab 9714 \tab 9719 \tab 9730 \tab 9740 \tab 9790 \tab 9830 \tab 9851 \tab 9852 \tab 9859 \tab 9891 \tab 9971 \tab 9990 \tab 9996 \tab 1639 \tab 1809 \tab 3100 \tab 350 \tab 3701 \tab 3930 \tab 4420 \tab 4513 \tab 5821 \tab 6001 \tab 6240 \tab 6320 \tab 7110 \tab 759 \tab 770 \tab 7790 \tab 790 \tab 8020 \tab 8120 \tab 8311 \tab 8330 \tab 8350 \tab 8420 \tab 8570 \tab 8590 \tab 8740 \tab 900 \tab 9250 \tab 9260 \tab 9560 \tab 9595 \tab 3800 \tab 3993 \tab 5920 \tab 7130 \tab 7230 \tab 769 \tab 8390 \tab 130 \tab 3801 \tab 3940 \tab 650 \tab 7280 \tab 8321 \tab 9570 \tab 131 \tab 4222 \tab 7940 \tab 9890 \tab 1229 \tab 330 \tab 810 \tab 820 \tab 8732 \tab 9240 \tab 9979 \tab 120 \tab 1924 \tab 2030 \tab 270 \tab 5910 \tab 5998 \tab 761 \tab 9810 \tab 7111 \tab 9559 \tab 7220 \tab 2020 \tab 5992 \tab 1929 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 87 \tab 2 \tab 13 \tab 6 \tab 10 \tab 8 \tab 27 \tab 6 \tab 2 \tab 8 \tab 4 \tab 2 \tab 5 \tab 1 \tab 1 \tab 4 \tab 2 \tab 4 \tab 8 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 4 \tab 7 \tab 6 \tab 5 \tab 1 \tab 9 \tab 100 \tab 1 \tab 3 \tab 5 \tab 3 \tab 1 \tab 1 \tab 6 \tab 6 \tab 4 \tab 1 \tab 2 \tab 7 \tab 60 \tab 4 \tab 9 \tab 32 \tab 5 \tab 26 \tab 3 \tab 5 \tab 4 \tab 3 \tab 2 \tab 1 \tab 5 \tab 3 \tab 5 \tab 6 \tab 9 \tab 3 \tab 8 \tab 9 \tab 11 \tab 21 \tab 13 \tab 2 \tab 1 \tab 12 \tab 1 \tab 1 \tab 1 \tab 19 \tab 2 \tab 1 \tab 3 \tab 3 \tab 10 \tab 8 \tab 2 \tab 5 \tab 1 \tab 6 \tab 64 \tab 3 \tab 5 \tab 1 \tab 2 \tab 3 \tab 9 \tab 1 \tab 6 \tab 5 \tab 20 \tab 7 \tab 6 \tab 9 \tab 10 \tab 6 \tab 10 \tab 20 \tab 4 \tab 24 \tab 1 \tab 14 \tab 5 \tab 2 \tab 9 \tab 1 \tab 4 \tab 5 \tab 6 \tab 4 \tab 7 \tab 4 \tab 1 \tab 5 \tab 25 \tab 1 \tab 6 \tab 15 \tab 1 \tab 6 \tab 6 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab 27 \tab 29 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 6 \tab 2 \tab 1 \tab 2 \tab 23 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 9 \tab 2 \tab 13 \tab 2 \tab 1 \tab 1 \tab 12 \tab 3 \tab 6 \tab 3 \tab 19 \tab 4 \tab 13 \tab 3 \tab 4 \tab 2 \tab 7 \tab 7 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab 5 \tab 1 \tab 1 \tab 5 \tab 4 \tab 5 \tab 2 \tab 3 \tab 2 \tab 9 \tab 4 \tab 1 \tab 7 \tab 34 \tab 1 \tab 1 \tab 9 \tab 4 \tab 7 \tab 6 \tab 1 \tab 3 \tab 1 \tab 6 \tab 1 \tab 10 \tab 21 \tab 1 \tab 2 \tab 1 \tab 17 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 96 \tab 2 \tab 12 \tab 13 \tab 18 \tab 15 \tab 23 \tab 5 \tab 2 \tab 9 \tab 1 \tab 5 \tab 2 \tab 2 \tab 1 \tab 4 \tab 2 \tab 2 \tab 5 \tab 2 \tab 2 \tab 3 \tab 2 \tab - \tab 4 \tab - \tab - \tab 7 \tab 10 \tab 4 \tab 10 \tab 4 \tab 5 \tab 108 \tab 2 \tab 3 \tab 5 \tab 2 \tab 2 \tab - \tab 2 \tab 3 \tab 13 \tab - \tab 4 \tab 2 \tab 82 \tab 2 \tab 7 \tab 30 \tab 5 \tab 15 \tab 2 \tab 2 \tab 3 \tab 6 \tab 8 \tab 1 \tab 7 \tab - \tab 4 \tab 2 \tab 4 \tab 8 \tab 7 \tab 10 \tab 8 \tab 27 \tab 9 \tab 2 \tab - \tab 18 \tab 1 \tab - \tab - \tab 33 \tab 2 \tab 1 \tab 15 \tab 5 \tab 5 \tab 5 \tab 1 \tab 8 \tab 8 \tab 3 \tab 38 \tab 4 \tab 4 \tab - \tab - \tab 1 \tab 16 \tab - \tab 3 \tab 5 \tab 17 \tab 11 \tab 2 \tab 4 \tab 17 \tab 10 \tab 11 \tab 13 \tab 6 \tab 24 \tab 6 \tab 11 \tab 11 \tab 1 \tab 7 \tab 2 \tab 2 \tab 6 \tab 8 \tab 1 \tab 17 \tab 6 \tab - \tab - \tab 24 \tab - \tab 3 \tab 16 \tab 1 \tab 4 \tab 12 \tab 5 \tab 3 \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 2 \tab 21 \tab 40 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 10 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 5 \tab 2 \tab 21 \tab 2 \tab 1 \tab 1 \tab 9 \tab 4 \tab 3 \tab 8 \tab 8 \tab 4 \tab 12 \tab 2 \tab 6 \tab 5 \tab 7 \tab 3 \tab - \tab - \tab 4 \tab 12 \tab 4 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 6 \tab - \tab 3 \tab 3 \tab 2 \tab 7 \tab 7 \tab 2 \tab - \tab 2 \tab 27 \tab 1 \tab - \tab 3 \tab 4 \tab 3 \tab 7 \tab 1 \tab - \tab 2 \tab 8 \tab 2 \tab 5 \tab 21 \tab 2 \tab - \tab 2 \tab 18 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 2 \tab 5 \tab 3 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 72 \tab 1 \tab 18 \tab 8 \tab 14 \tab 14 \tab 34 \tab 2 \tab 1 \tab 9 \tab 4 \tab 3 \tab 2 \tab 2 \tab 2 \tab 5 \tab 5 \tab 6 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 6 \tab 3 \tab 1 \tab 13 \tab 10 \tab 4 \tab 2 \tab 3 \tab 3 \tab 72 \tab 1 \tab 2 \tab 8 \tab 3 \tab - \tab - \tab 2 \tab 3 \tab 15 \tab 1 \tab 2 \tab 3 \tab 63 \tab 6 \tab 4 \tab 28 \tab 3 \tab 21 \tab 2 \tab 2 \tab 7 \tab 6 \tab 3 \tab - \tab 4 \tab - \tab 9 \tab 2 \tab 3 \tab 2 \tab 8 \tab 6 \tab 6 \tab 22 \tab 11 \tab 2 \tab 2 \tab 11 \tab - \tab 2 \tab - \tab 23 \tab 2 \tab - \tab 12 \tab 6 \tab 4 \tab 14 \tab 1 \tab 8 \tab 8 \tab 3 \tab 40 \tab 1 \tab 4 \tab 2 \tab - \tab - \tab 9 \tab - \tab 5 \tab 7 \tab 21 \tab 2 \tab 4 \tab 3 \tab 15 \tab 3 \tab 5 \tab 15 \tab 3 \tab 13 \tab 4 \tab 10 \tab 3 \tab 1 \tab 11 \tab - \tab 2 \tab 3 \tab 6 \tab 3 \tab 6 \tab 2 \tab 2 \tab 5 \tab 23 \tab 1 \tab 4 \tab 16 \tab - \tab 5 \tab 11 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 18 \tab 30 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 3 \tab 2 \tab - \tab 2 \tab 2 \tab 3 \tab - \tab 14 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 2 \tab 21 \tab 4 \tab - \tab - \tab 6 \tab 2 \tab 2 \tab 4 \tab 15 \tab 1 \tab 11 \tab 3 \tab 4 \tab 2 \tab 5 \tab 3 \tab - \tab - \tab 7 \tab 4 \tab 6 \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 5 \tab 1 \tab 16 \tab 3 \tab - \tab - \tab 43 \tab - \tab 1 \tab 7 \tab 7 \tab 6 \tab 13 \tab 1 \tab - \tab 2 \tab 5 \tab 1 \tab 6 \tab 29 \tab 1 \tab 1 \tab 2 \tab 11 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 6 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 99 \tab 1 \tab 19 \tab 9 \tab 8 \tab 17 \tab 38 \tab 8 \tab - \tab 7 \tab 6 \tab 3 \tab - \tab 1 \tab - \tab 7 \tab 3 \tab 4 \tab 4 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 9 \tab 4 \tab 5 \tab 12 \tab 3 \tab 7 \tab 82 \tab 5 \tab 1 \tab 5 \tab 4 \tab - \tab - \tab 3 \tab 3 \tab 13 \tab - \tab 4 \tab 3 \tab 64 \tab 8 \tab 10 \tab 35 \tab 5 \tab 24 \tab 1 \tab 3 \tab 4 \tab 6 \tab 9 \tab - \tab 8 \tab - \tab 7 \tab 4 \tab 4 \tab 3 \tab 4 \tab 6 \tab 5 \tab 14 \tab 8 \tab 7 \tab 5 \tab 12 \tab 1 \tab 1 \tab - \tab 32 \tab 5 \tab - \tab 14 \tab 4 \tab 5 \tab 6 \tab 1 \tab 11 \tab 8 \tab 4 \tab 38 \tab 1 \tab 8 \tab - \tab - \tab - \tab 4 \tab 1 \tab 4 \tab 4 \tab 27 \tab 1 \tab 5 \tab 5 \tab 24 \tab 4 \tab 5 \tab 16 \tab 6 \tab 21 \tab 3 \tab 5 \tab 11 \tab 1 \tab 10 \tab 2 \tab 2 \tab 5 \tab 12 \tab 4 \tab 12 \tab 2 \tab 1 \tab 3 \tab 32 \tab 9 \tab 4 \tab 14 \tab - \tab 5 \tab 7 \tab 5 \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 29 \tab 35 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab 3 \tab 3 \tab - \tab 23 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 6 \tab - \tab 27 \tab - \tab 2 \tab 3 \tab 9 \tab 1 \tab 3 \tab 6 \tab 7 \tab 5 \tab 9 \tab 5 \tab 4 \tab 1 \tab 5 \tab 3 \tab 2 \tab - \tab 8 \tab 5 \tab 2 \tab 1 \tab 3 \tab - \tab 3 \tab 4 \tab 2 \tab 7 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 16 \tab 5 \tab - \tab 2 \tab 38 \tab - \tab - \tab 7 \tab 4 \tab 8 \tab 9 \tab - \tab 1 \tab 2 \tab 7 \tab 1 \tab 2 \tab 18 \tab 3 \tab 3 \tab 2 \tab 9 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 4 \tab 1 \tab 6 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 80 \tab - \tab 18 \tab 18 \tab 7 \tab 13 \tab 30 \tab 4 \tab 4 \tab 9 \tab 4 \tab 4 \tab 3 \tab - \tab 5 \tab 2 \tab 2 \tab 6 \tab 6 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 4 \tab 3 \tab 4 \tab 11 \tab 10 \tab 2 \tab 6 \tab 3 \tab 6 \tab 93 \tab 1 \tab 3 \tab - \tab 3 \tab 2 \tab - \tab 4 \tab 4 \tab 16 \tab - \tab 2 \tab 3 \tab 60 \tab 13 \tab 10 \tab 36 \tab 6 \tab 26 \tab 3 \tab 6 \tab 6 \tab 1 \tab 3 \tab 1 \tab 12 \tab - \tab 5 \tab 3 \tab 6 \tab 1 \tab 7 \tab 13 \tab 7 \tab 19 \tab 10 \tab 5 \tab 4 \tab 11 \tab 1 \tab 2 \tab - \tab 50 \tab 12 \tab 2 \tab 20 \tab - \tab 15 \tab 11 \tab 4 \tab 13 \tab 12 \tab 5 \tab 28 \tab 2 \tab 6 \tab 1 \tab 5 \tab - \tab 15 \tab - \tab 2 \tab 6 \tab 22 \tab - \tab 3 \tab 2 \tab 27 \tab 7 \tab 10 \tab 15 \tab 5 \tab 19 \tab 2 \tab 12 \tab 3 \tab 1 \tab 20 \tab 1 \tab 3 \tab 6 \tab 10 \tab 2 \tab 9 \tab 3 \tab 3 \tab - \tab 28 \tab 1 \tab 7 \tab 16 \tab - \tab 7 \tab 8 \tab 8 \tab 2 \tab - \tab 1 \tab 6 \tab - \tab 1 \tab 2 \tab 28 \tab 31 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab - \tab 3 \tab 3 \tab 2 \tab - \tab 22 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 8 \tab 4 \tab 21 \tab 2 \tab - \tab - \tab 10 \tab 2 \tab 4 \tab - \tab 14 \tab 2 \tab 11 \tab 5 \tab 5 \tab 1 \tab 12 \tab 3 \tab 1 \tab - \tab 8 \tab 6 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 7 \tab 2 \tab 5 \tab 6 \tab 3 \tab 2 \tab 20 \tab - \tab - \tab 3 \tab 28 \tab - \tab 1 \tab 3 \tab 6 \tab 4 \tab 8 \tab - \tab 1 \tab 1 \tab 13 \tab - \tab 6 \tab 25 \tab - \tab 1 \tab - \tab 21 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 9 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 139 \tab 3 \tab 36 \tab 31 \tab 19 \tab 25 \tab 68 \tab 6 \tab 5 \tab 14 \tab 12 \tab 7 \tab 6 \tab 2 \tab 6 \tab 10 \tab 4 \tab 13 \tab 10 \tab 8 \tab 4 \tab 5 \tab 1 \tab - \tab 4 \tab 5 \tab 1 \tab 29 \tab 17 \tab 4 \tab 7 \tab 6 \tab 5 \tab 161 \tab 5 \tab 5 \tab 11 \tab 6 \tab 3 \tab - \tab 8 \tab 5 \tab 28 \tab 1 \tab 9 \tab 5 \tab 112 \tab 11 \tab 15 \tab 80 \tab 10 \tab 47 \tab 10 \tab 5 \tab 13 \tab 5 \tab 5 \tab - \tab 13 \tab 4 \tab 20 \tab 10 \tab 14 \tab 3 \tab 22 \tab 8 \tab 14 \tab 27 \tab 20 \tab 7 \tab 8 \tab 41 \tab 1 \tab 4 \tab - \tab 78 \tab 12 \tab 2 \tab 34 \tab 8 \tab 17 \tab 18 \tab 9 \tab 12 \tab 18 \tab 8 \tab 60 \tab 7 \tab 9 \tab 3 \tab - \tab 2 \tab 16 \tab 2 \tab 16 \tab 5 \tab 30 \tab 1 \tab 9 \tab 4 \tab 51 \tab 14 \tab 20 \tab 38 \tab 6 \tab 42 \tab 9 \tab 25 \tab 8 \tab 2 \tab 25 \tab 2 \tab 7 \tab 13 \tab 24 \tab 6 \tab 24 \tab 2 \tab 5 \tab 7 \tab 67 \tab 5 \tab 10 \tab 26 \tab 1 \tab 9 \tab 8 \tab 14 \tab 4 \tab - \tab 2 \tab 4 \tab - \tab 3 \tab - \tab 37 \tab 63 \tab - \tab 3 \tab 3 \tab 6 \tab 4 \tab 1 \tab 3 \tab 2 \tab 4 \tab 5 \tab 9 \tab - \tab 2 \tab 17 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 33 \tab 3 \tab 36 \tab 4 \tab 1 \tab 1 \tab 17 \tab 5 \tab 17 \tab 11 \tab 21 \tab 10 \tab 23 \tab 8 \tab 15 \tab 1 \tab 16 \tab 8 \tab 1 \tab - \tab 15 \tab 16 \tab 8 \tab - \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 9 \tab 2 \tab 8 \tab 3 \tab 3 \tab 12 \tab 21 \tab 5 \tab 6 \tab 6 \tab 50 \tab - \tab 1 \tab 7 \tab 16 \tab 13 \tab 19 \tab - \tab 1 \tab 3 \tab 23 \tab 1 \tab 12 \tab 56 \tab 7 \tab 1 \tab 1 \tab 26 \tab 1 \tab 5 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 4 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab 4 \tab 4 \tab 5 \tab 5 \tab 2 \tab - \tab 3 \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 123 \tab 3 \tab 26 \tab 18 \tab 17 \tab 32 \tab 62 \tab 9 \tab 6 \tab 13 \tab 11 \tab 8 \tab 10 \tab 2 \tab 1 \tab 5 \tab 6 \tab 10 \tab 14 \tab 7 \tab - \tab 4 \tab 4 \tab - \tab 7 \tab 3 \tab 1 \tab 22 \tab 19 \tab 13 \tab 23 \tab 3 \tab 13 \tab 238 \tab 6 \tab 10 \tab 13 \tab 3 \tab - \tab 1 \tab 6 \tab 9 \tab 13 \tab 1 \tab 7 \tab 4 \tab 82 \tab 8 \tab 7 \tab 53 \tab 16 \tab 43 \tab 5 \tab 3 \tab 10 \tab 1 \tab 5 \tab - \tab 14 \tab - \tab 8 \tab 10 \tab 7 \tab 4 \tab 13 \tab 18 \tab 12 \tab 33 \tab 30 \tab 9 \tab 3 \tab 29 \tab 1 \tab 3 \tab - \tab 70 \tab 16 \tab - \tab 10 \tab 13 \tab 15 \tab 11 \tab 5 \tab 11 \tab 17 \tab 5 \tab 89 \tab 5 \tab 10 \tab 1 \tab 2 \tab 5 \tab 37 \tab 1 \tab 18 \tab 5 \tab 35 \tab 3 \tab 7 \tab 1 \tab 33 \tab 7 \tab 19 \tab 33 \tab 10 \tab 53 \tab 6 \tab 17 \tab 11 \tab - \tab 27 \tab - \tab 7 \tab 9 \tab 21 \tab 6 \tab - \tab 10 \tab 5 \tab 5 \tab 54 \tab 5 \tab 15 \tab 15 \tab 1 \tab 12 \tab 11 \tab 14 \tab 5 \tab - \tab 1 \tab 2 \tab 3 \tab 4 \tab 4 \tab 59 \tab 67 \tab 1 \tab 3 \tab 1 \tab 4 \tab 5 \tab 2 \tab 3 \tab 1 \tab 2 \tab 5 \tab 1 \tab 1 \tab - \tab 23 \tab 5 \tab 4 \tab 1 \tab 1 \tab 1 \tab 31 \tab 3 \tab 39 \tab 2 \tab - \tab 1 \tab 22 \tab 5 \tab 13 \tab 7 \tab 36 \tab 9 \tab 24 \tab 9 \tab 12 \tab - \tab 10 \tab 3 \tab - \tab - \tab 17 \tab 15 \tab 11 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 9 \tab 2 \tab 8 \tab 3 \tab 2 \tab 7 \tab 19 \tab 3 \tab 4 \tab 1 \tab 60 \tab 1 \tab 1 \tab 19 \tab 13 \tab 10 \tab 13 \tab - \tab 2 \tab 1 \tab 21 \tab 2 \tab 15 \tab 53 \tab 4 \tab 3 \tab 2 \tab 38 \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 4 \tab 1 \tab 5 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 154 \tab 2 \tab 27 \tab 23 \tab 18 \tab 25 \tab 48 \tab 6 \tab 9 \tab 22 \tab 7 \tab 5 \tab 9 \tab 1 \tab 2 \tab 15 \tab 7 \tab 5 \tab 13 \tab 8 \tab 1 \tab 2 \tab 2 \tab 1 \tab 8 \tab 9 \tab - \tab 19 \tab 17 \tab 9 \tab 12 \tab 4 \tab 14 \tab 253 \tab 1 \tab 4 \tab 10 \tab 8 \tab 3 \tab - \tab 4 \tab 8 \tab 6 \tab - \tab 8 \tab 2 \tab 93 \tab 6 \tab 7 \tab 54 \tab 9 \tab 34 \tab 5 \tab 3 \tab 11 \tab 1 \tab 4 \tab - \tab 10 \tab 1 \tab 11 \tab 6 \tab 10 \tab 3 \tab 16 \tab 13 \tab 12 \tab 48 \tab 20 \tab 2 \tab 5 \tab 24 \tab - \tab 8 \tab - \tab 69 \tab 7 \tab - \tab 33 \tab 13 \tab 22 \tab 21 \tab 10 \tab 17 \tab 15 \tab 8 \tab 79 \tab 3 \tab 12 \tab 1 \tab - \tab 1 \tab 26 \tab 2 \tab 11 \tab 5 \tab 35 \tab 5 \tab 9 \tab 6 \tab 39 \tab 8 \tab 14 \tab 27 \tab 11 \tab 43 \tab 7 \tab 12 \tab 12 \tab 2 \tab 19 \tab - \tab 4 \tab 14 \tab 15 \tab 2 \tab 6 \tab 11 \tab 4 \tab 1 \tab 56 \tab 6 \tab 9 \tab 18 \tab 1 \tab 16 \tab 6 \tab 8 \tab 3 \tab - \tab 1 \tab 8 \tab - \tab 4 \tab 3 \tab 44 \tab 79 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab 7 \tab - \tab 7 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 31 \tab 4 \tab 43 \tab - \tab 1 \tab - \tab 21 \tab 5 \tab 25 \tab 7 \tab 41 \tab 4 \tab 13 \tab 3 \tab 8 \tab 4 \tab 20 \tab 5 \tab - \tab - \tab 9 \tab 13 \tab 7 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab 7 \tab 1 \tab 10 \tab 1 \tab 2 \tab 6 \tab 15 \tab 7 \tab 3 \tab 2 \tab 61 \tab 1 \tab 1 \tab 10 \tab 8 \tab 11 \tab 20 \tab 1 \tab 2 \tab 1 \tab 15 \tab 1 \tab 10 \tab 52 \tab 5 \tab 2 \tab - \tab 31 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 5 \tab - \tab 2 \tab 3 \tab 4 \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 5 \tab - \tab 2 \tab 2 \tab 4 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 157 \tab - \tab 40 \tab 17 \tab 26 \tab 29 \tab 65 \tab 11 \tab 8 \tab 12 \tab 8 \tab 3 \tab 6 \tab - \tab 3 \tab 9 \tab 4 \tab 7 \tab 8 \tab 5 \tab - \tab 3 \tab 1 \tab 5 \tab 5 \tab 8 \tab - \tab 29 \tab 11 \tab 6 \tab 9 \tab 6 \tab 21 \tab 287 \tab - \tab 4 \tab 13 \tab 2 \tab 4 \tab 1 \tab 10 \tab 2 \tab 3 \tab - \tab 5 \tab 5 \tab 68 \tab 8 \tab 11 \tab 55 \tab 5 \tab 29 \tab 9 \tab 2 \tab 22 \tab 3 \tab 4 \tab 1 \tab 7 \tab 6 \tab 14 \tab 7 \tab 6 \tab 2 \tab 16 \tab 7 \tab 9 \tab 30 \tab 19 \tab 2 \tab 4 \tab 34 \tab - \tab 2 \tab - \tab 59 \tab 7 \tab - \tab 14 \tab 7 \tab 22 \tab 25 \tab 9 \tab 9 \tab 21 \tab 11 \tab 85 \tab 3 \tab 9 \tab 1 \tab - \tab 2 \tab 19 \tab 3 \tab 16 \tab 6 \tab 23 \tab 2 \tab 12 \tab 13 \tab 35 \tab 8 \tab 25 \tab 33 \tab 18 \tab 54 \tab 2 \tab 23 \tab 9 \tab - \tab 17 \tab 1 \tab 3 \tab 13 \tab 19 \tab 2 \tab 7 \tab 2 \tab 1 \tab 3 \tab 53 \tab - \tab 8 \tab 21 \tab 3 \tab 10 \tab 9 \tab 13 \tab 4 \tab - \tab 1 \tab 9 \tab 1 \tab 1 \tab 1 \tab 33 \tab 78 \tab - \tab 3 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 4 \tab 3 \tab 19 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 38 \tab 1 \tab 44 \tab 2 \tab - \tab 1 \tab 17 \tab 6 \tab 9 \tab 10 \tab 28 \tab 6 \tab 19 \tab 3 \tab 15 \tab - \tab 17 \tab 3 \tab 2 \tab 1 \tab 16 \tab 18 \tab 2 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 7 \tab 1 \tab 10 \tab 3 \tab 1 \tab 1 \tab 22 \tab 8 \tab 1 \tab 5 \tab 53 \tab 2 \tab - \tab 18 \tab 11 \tab 12 \tab 13 \tab - \tab 2 \tab 3 \tab 21 \tab - \tab 9 \tab 45 \tab 7 \tab - \tab - \tab 27 \tab - \tab 4 \tab 1 \tab - \tab - \tab 2 \tab - \tab 4 \tab 2 \tab 3 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 4 \tab 2 \tab 3 \tab 4 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab 8 \tab 1 \tab 3 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 117 \tab 4 \tab 26 \tab 23 \tab 15 \tab 32 \tab 60 \tab 6 \tab 5 \tab 17 \tab 11 \tab 7 \tab 4 \tab 1 \tab 2 \tab 11 \tab 4 \tab 7 \tab 13 \tab 5 \tab - \tab 6 \tab 3 \tab - \tab 6 \tab 8 \tab 1 \tab 25 \tab 19 \tab 4 \tab 10 \tab 5 \tab 6 \tab 259 \tab 5 \tab 6 \tab 6 \tab 7 \tab 1 \tab - \tab 7 \tab 3 \tab 8 \tab 1 \tab 4 \tab 6 \tab 58 \tab 10 \tab 11 \tab 48 \tab 7 \tab 31 \tab 7 \tab 2 \tab 19 \tab - \tab 4 \tab - \tab 24 \tab - \tab 17 \tab 4 \tab 8 \tab 8 \tab 15 \tab 8 \tab 11 \tab 36 \tab 24 \tab 2 \tab 2 \tab 29 \tab - \tab 5 \tab - \tab 61 \tab 14 \tab - \tab 34 \tab 9 \tab 11 \tab 20 \tab 6 \tab 11 \tab 18 \tab 5 \tab 81 \tab 2 \tab 12 \tab 3 \tab 1 \tab 1 \tab 20 \tab 3 \tab 15 \tab 8 \tab 38 \tab 5 \tab 5 \tab 12 \tab 38 \tab 7 \tab 14 \tab 42 \tab 10 \tab 49 \tab 6 \tab 20 \tab 9 \tab - \tab 17 \tab - \tab 11 \tab 10 \tab 23 \tab 4 \tab 11 \tab 11 \tab 1 \tab 4 \tab 65 \tab - \tab 12 \tab 8 \tab - \tab 18 \tab 15 \tab 15 \tab 2 \tab - \tab 4 \tab 4 \tab 2 \tab 3 \tab 3 \tab 53 \tab 50 \tab - \tab 3 \tab - \tab 4 \tab - \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 9 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab 49 \tab 5 \tab 40 \tab 2 \tab - \tab 2 \tab 13 \tab 2 \tab 12 \tab 7 \tab 23 \tab 6 \tab 25 \tab 4 \tab 15 \tab - \tab 14 \tab 6 \tab 3 \tab 1 \tab 9 \tab 8 \tab 5 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 13 \tab 1 \tab 9 \tab 3 \tab 1 \tab 5 \tab 10 \tab 5 \tab - \tab 2 \tab 58 \tab - \tab - \tab 15 \tab 12 \tab 8 \tab 14 \tab - \tab 2 \tab 3 \tab 18 \tab 1 \tab 13 \tab 52 \tab 7 \tab 2 \tab 1 \tab 22 \tab - \tab 7 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 5 \tab 4 \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 9 \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab 6 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 5 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2765 \cr
   2004 \tab 153 \tab 7 \tab 37 \tab 24 \tab 25 \tab 32 \tab 71 \tab 10 \tab 2 \tab 17 \tab 19 \tab 1 \tab 10 \tab 2 \tab 3 \tab 5 \tab 3 \tab 2 \tab 18 \tab 5 \tab - \tab 4 \tab 3 \tab - \tab 6 \tab 10 \tab - \tab 32 \tab 18 \tab 15 \tab 29 \tab 3 \tab 13 \tab 271 \tab 1 \tab 7 \tab 13 \tab 11 \tab 3 \tab - \tab 2 \tab 9 \tab 13 \tab 1 \tab 7 \tab 4 \tab 52 \tab 4 \tab 7 \tab 49 \tab 7 \tab 40 \tab 12 \tab 3 \tab 11 \tab 1 \tab 5 \tab 2 \tab 16 \tab - \tab 10 \tab 3 \tab 7 \tab 5 \tab 9 \tab 5 \tab 18 \tab 27 \tab 17 \tab 5 \tab 7 \tab 25 \tab - \tab 3 \tab - \tab 73 \tab 8 \tab 1 \tab 30 \tab 12 \tab 19 \tab 18 \tab 13 \tab 13 \tab 15 \tab 8 \tab 72 \tab 4 \tab 3 \tab - \tab - \tab 1 \tab 16 \tab 2 \tab 21 \tab 4 \tab 19 \tab 4 \tab 10 \tab 11 \tab 43 \tab 15 \tab 15 \tab 27 \tab 15 \tab 45 \tab - \tab 14 \tab 7 \tab - \tab 14 \tab 1 \tab 10 \tab 13 \tab 12 \tab 4 \tab 10 \tab 6 \tab 1 \tab 5 \tab 65 \tab 7 \tab 14 \tab 16 \tab - \tab 18 \tab 12 \tab 12 \tab 4 \tab - \tab 3 \tab 4 \tab 2 \tab 3 \tab 3 \tab 33 \tab 62 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 4 \tab 4 \tab 3 \tab 3 \tab 13 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 48 \tab 2 \tab 43 \tab - \tab - \tab - \tab 18 \tab 5 \tab 8 \tab 5 \tab 22 \tab 3 \tab 20 \tab 4 \tab 9 \tab - \tab 12 \tab 4 \tab 5 \tab 1 \tab 9 \tab 12 \tab 10 \tab - \tab - \tab 1 \tab 4 \tab 2 \tab 1 \tab 9 \tab 3 \tab 3 \tab 2 \tab 2 \tab 4 \tab 9 \tab 6 \tab - \tab 1 \tab 69 \tab - \tab - \tab 12 \tab 11 \tab 7 \tab 11 \tab - \tab - \tab 4 \tab 19 \tab 3 \tab 14 \tab 51 \tab 7 \tab 3 \tab - \tab 22 \tab 1 \tab 5 \tab 5 \tab 1 \tab - \tab 4 \tab - \tab - \tab 3 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 2 \tab 3 \tab 7 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 9 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2812 \cr
   2006 \tab 240 \tab 2 \tab 38 \tab 37 \tab 39 \tab 84 \tab 87 \tab 13 \tab 2 \tab 17 \tab 10 \tab 14 \tab 11 \tab 5 \tab 4 \tab 19 \tab 8 \tab 10 \tab 28 \tab 1 \tab 1 \tab 6 \tab 3 \tab 2 \tab 10 \tab 5 \tab 7 \tab 44 \tab 29 \tab 22 \tab 16 \tab 8 \tab 12 \tab 365 \tab 3 \tab 12 \tab 14 \tab 13 \tab 4 \tab - \tab 16 \tab 16 \tab 22 \tab 2 \tab 12 \tab 7 \tab 97 \tab 3 \tab 14 \tab 70 \tab 27 \tab 55 \tab 5 \tab 8 \tab 21 \tab 5 \tab 8 \tab 3 \tab 30 \tab 1 \tab 17 \tab 8 \tab 5 \tab 4 \tab 14 \tab 16 \tab 22 \tab 56 \tab 29 \tab 15 \tab 5 \tab 50 \tab - \tab 4 \tab - \tab 80 \tab 9 \tab - \tab 38 \tab 11 \tab 19 \tab 37 \tab 12 \tab 23 \tab 46 \tab 15 \tab 119 \tab 2 \tab 4 \tab 3 \tab 1 \tab 6 \tab 26 \tab 1 \tab 22 \tab 9 \tab 61 \tab 4 \tab 12 \tab 9 \tab 58 \tab 18 \tab 26 \tab 74 \tab 7 \tab 70 \tab 7 \tab 48 \tab 12 \tab 3 \tab 33 \tab 5 \tab 10 \tab 25 \tab 25 \tab 7 \tab 28 \tab 28 \tab 3 \tab 4 \tab 67 \tab 8 \tab 25 \tab 22 \tab 5 \tab 47 \tab 19 \tab 18 \tab 6 \tab - \tab 2 \tab 14 \tab 5 \tab 6 \tab 2 \tab 82 \tab 89 \tab 2 \tab 8 \tab 2 \tab 2 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 5 \tab 6 \tab 8 \tab 2 \tab 13 \tab 7 \tab 5 \tab 1 \tab 7 \tab - \tab 76 \tab 3 \tab 72 \tab 5 \tab - \tab 2 \tab 25 \tab 3 \tab 28 \tab 29 \tab 33 \tab 6 \tab 31 \tab 20 \tab 12 \tab 3 \tab 21 \tab 6 \tab 1 \tab - \tab 18 \tab 18 \tab 5 \tab 3 \tab 1 \tab 4 \tab 2 \tab 2 \tab - \tab 10 \tab 2 \tab 15 \tab 5 \tab 2 \tab 8 \tab 14 \tab 8 \tab 2 \tab 12 \tab 106 \tab 1 \tab - \tab 29 \tab 8 \tab 23 \tab 26 \tab - \tab 5 \tab 4 \tab 37 \tab 2 \tab 13 \tab 85 \tab 7 \tab 4 \tab 1 \tab 46 \tab 3 \tab 1 \tab 7 \tab 1 \tab 1 \tab 5 \tab - \tab 12 \tab - \tab 2 \tab 2 \tab 2 \tab 3 \tab - \tab - \tab 6 \tab 4 \tab 15 \tab - \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 9 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 7 \tab 3 \tab 1 \tab - \tab 9 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 5 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 102 \tab 2 \tab 25 \tab 18 \tab 19 \tab 22 \tab 35 \tab 11 \tab 1 \tab 18 \tab 4 \tab - \tab 7 \tab - \tab 2 \tab 7 \tab 4 \tab 3 \tab 16 \tab 4 \tab - \tab 3 \tab 1 \tab - \tab 3 \tab 6 \tab - \tab 19 \tab 14 \tab 9 \tab 10 \tab 3 \tab 9 \tab 177 \tab 2 \tab 5 \tab 4 \tab 3 \tab 2 \tab - \tab 13 \tab 8 \tab 13 \tab 1 \tab 7 \tab 6 \tab 36 \tab 4 \tab 5 \tab 33 \tab 3 \tab 14 \tab 4 \tab 3 \tab 9 \tab 1 \tab 4 \tab - \tab 9 \tab - \tab 5 \tab 7 \tab 5 \tab 5 \tab 13 \tab 5 \tab 20 \tab 27 \tab 14 \tab 9 \tab 2 \tab 24 \tab - \tab 1 \tab - \tab 22 \tab 6 \tab 1 \tab 15 \tab 5 \tab 16 \tab 12 \tab 8 \tab 8 \tab 3 \tab 3 \tab 55 \tab - \tab 7 \tab - \tab - \tab 1 \tab 17 \tab - \tab 13 \tab 5 \tab 29 \tab 2 \tab 8 \tab 3 \tab 36 \tab 7 \tab 14 \tab 23 \tab 8 \tab 30 \tab 2 \tab 14 \tab 6 \tab - \tab 15 \tab 1 \tab 4 \tab 9 \tab 15 \tab 8 \tab 10 \tab 18 \tab 3 \tab 1 \tab 35 \tab 3 \tab 5 \tab 12 \tab 2 \tab 16 \tab 14 \tab 8 \tab 3 \tab - \tab 1 \tab 4 \tab 2 \tab - \tab - \tab 35 \tab 43 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 6 \tab 4 \tab 3 \tab - \tab 8 \tab 1 \tab 5 \tab - \tab 2 \tab 1 \tab 24 \tab 1 \tab 41 \tab 2 \tab - \tab - \tab 13 \tab 2 \tab 13 \tab 6 \tab 6 \tab 7 \tab 10 \tab 9 \tab 2 \tab 2 \tab 5 \tab 3 \tab 5 \tab - \tab 6 \tab 14 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 6 \tab 3 \tab 6 \tab - \tab 2 \tab 1 \tab 9 \tab 5 \tab 1 \tab 3 \tab 62 \tab - \tab - \tab 8 \tab 4 \tab 5 \tab 5 \tab 3 \tab 2 \tab 1 \tab 21 \tab 1 \tab 7 \tab 53 \tab 3 \tab - \tab - \tab 22 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 5 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2023 \cr
   2010 \tab 154 \tab 2 \tab 16 \tab 13 \tab 18 \tab 23 \tab 47 \tab 13 \tab 2 \tab 11 \tab 5 \tab 6 \tab 6 \tab 2 \tab 1 \tab 5 \tab 2 \tab 4 \tab 9 \tab 5 \tab - \tab 3 \tab 1 \tab - \tab 8 \tab 2 \tab 3 \tab 13 \tab 19 \tab 18 \tab 13 \tab 1 \tab 8 \tab 174 \tab 1 \tab 3 \tab 6 \tab 8 \tab - \tab 1 \tab 3 \tab 10 \tab 11 \tab 1 \tab 5 \tab 3 \tab 42 \tab 4 \tab 4 \tab 45 \tab 3 \tab 19 \tab 4 \tab - \tab 4 \tab - \tab 2 \tab - \tab 17 \tab - \tab 3 \tab 5 \tab 3 \tab 4 \tab 6 \tab 8 \tab 24 \tab 27 \tab 7 \tab 2 \tab - \tab 38 \tab - \tab 1 \tab - \tab 44 \tab 3 \tab - \tab 13 \tab 7 \tab 12 \tab 12 \tab 5 \tab 11 \tab 11 \tab 3 \tab 72 \tab 1 \tab 5 \tab - \tab 1 \tab 3 \tab 14 \tab 2 \tab 8 \tab 4 \tab 28 \tab 2 \tab 5 \tab 4 \tab 40 \tab 14 \tab 10 \tab 20 \tab 9 \tab 45 \tab 2 \tab 17 \tab 4 \tab 1 \tab 17 \tab 2 \tab 2 \tab 13 \tab 11 \tab 3 \tab 15 \tab 26 \tab 2 \tab 5 \tab 32 \tab 3 \tab 2 \tab 8 \tab 1 \tab 15 \tab 8 \tab 14 \tab 3 \tab - \tab 1 \tab 6 \tab 3 \tab 3 \tab 1 \tab 25 \tab 42 \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab 11 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 23 \tab 3 \tab 18 \tab 1 \tab - \tab - \tab 14 \tab 3 \tab 4 \tab 10 \tab 11 \tab - \tab 6 \tab 2 \tab 5 \tab 2 \tab 7 \tab 2 \tab - \tab 1 \tab 8 \tab 13 \tab 6 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab 8 \tab 3 \tab 1 \tab 2 \tab 6 \tab 3 \tab 1 \tab 4 \tab 42 \tab - \tab - \tab 9 \tab 4 \tab 3 \tab 14 \tab 1 \tab 3 \tab 2 \tab 16 \tab 1 \tab 6 \tab 36 \tab 3 \tab - \tab 1 \tab 18 \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab 5 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 40951 \tab 31 \tab 351 \tab 258 \tab 253 \tab 371 \tab 695 \tab 110 \tab 49 \tab 183 \tab 106 \tab 68 \tab 81 \tab 21 \tab 33 \tab 108 \tab 56 \tab 83 \tab 159 \tab 56 \tab 11 \tab 47 \tab 25 \tab 10 \tab 73 \tab 66 \tab 20 \tab 276 \tab 204 \tab 121 \tab 164 \tab 53 \tab 131 \tab 2640 \tab 34 \tab 68 \tab 113 \tab 76 \tab 25 \tab 4 \tab 86 \tab 89 \tab 178 \tab 10 \tab 78 \tab 60 \tab 969 \tab 91 \tab 121 \tab 648 \tab 111 \tab 424 \tab 72 \tab 47 \tab 144 \tab 39 \tab 66 \tab 9 \tab 176 \tab 15 \tab 135 \tab 77 \tab 91 \tab 55 \tab 158 \tab 132 \tab 179 \tab 414 \tab 231 \tab 71 \tab 48 \tab 358 \tab 6 \tab 37 \tab 1 \tab 713 \tab 105 \tab 8 \tab 285 \tab 103 \tab 192 \tab 218 \tab 86 \tab 160 \tab 201 \tab 87 \tab 920 \tab 38 \tab 98 \tab 16 \tab 12 \tab 26 \tab 244 \tab 18 \tab 160 \tab 78 \tab 405 \tab 49 \tab 97 \tab 86 \tab 466 \tab 128 \tab 198 \tab 396 \tab 118 \tab 532 \tab 57 \tab 242 \tab 111 \tab 14 \tab 241 \tab 18 \tab 71 \tab 144 \tab 207 \tab 56 \tab 162 \tab 131 \tab 32 \tab 48 \tab 626 \tab 49 \tab 124 \tab 223 \tab 16 \tab 188 \tab 146 \tab 140 \tab 43 \tab 1 \tab 20 \tab 71 \tab 19 \tab 35 \tab 26 \tab 524 \tab 738 \tab 14 \tab 36 \tab 10 \tab 32 \tab 26 \tab 14 \tab 32 \tab 19 \tab 16 \tab 49 \tab 47 \tab 43 \tab 15 \tab 212 \tab 26 \tab 33 \tab 10 \tab 24 \tab 15 \tab 389 \tab 35 \tab 479 \tab 28 \tab 6 \tab 12 \tab 206 \tab 48 \tab 147 \tab 113 \tab 284 \tab 67 \tab 227 \tab 80 \tab 116 \tab 23 \tab 158 \tab 59 \tab 21 \tab 5 \tab 135 \tab 157 \tab 74 \tab 22 \tab 13 \tab 21 \tab 31 \tab 12 \tab 14 \tab 100 \tab 22 \tab 94 \tab 38 \tab 31 \tab 61 \tab 193 \tab 64 \tab 19 \tab 50 \tab 731 \tab 7 \tab 6 \tab 156 \tab 112 \tab 120 \tab 178 \tab 8 \tab 24 \tab 30 \tab 230 \tab 17 \tab 128 \tab 597 \tab 57 \tab 22 \tab 13 \tab 328 \tab 9 \tab 36 \tab 32 \tab 11 \tab 4 \tab 35 \tab 3 \tab 42 \tab 18 \tab 20 \tab 6 \tab 29 \tab 7 \tab 9 \tab 7 \tab 37 \tab 21 \tab 77 \tab 17 \tab 16 \tab 3 \tab 19 \tab 2 \tab 6 \tab 25 \tab 17 \tab 23 \tab 62 \tab 11 \tab 12 \tab 11 \tab 10 \tab 2 \tab 11 \tab 13 \tab 9 \tab 4 \tab 31 \tab 10 \tab 13 \tab 1 \tab 31 \tab 11 \tab 8 \tab 1 \tab 5 \tab 8 \tab 5 \tab 7 \tab 10 \tab 5 \tab 9 \tab 10 \tab 2 \tab 4 \tab 1 \tab 7 \tab 4 \tab 8 \tab 5 \tab 3 \tab 6 \tab 6 \tab 16 \tab 5 \tab 2 \tab 6 \tab 1 \tab 4 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
