% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_596.R
\name{isco88}
\alias{isco88}
\title{isco88}
\source{
General Social Survey https://gss.norc.org
}
\description{
Respondent's occupation, 1980 census & 1988 isco code
}
\details{
Question 1642. Respondent's Occupation, 1988 Census
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 110 \tab 1210 \tab 1220 \tab 1221 \tab 1223 \tab 1225 \tab 1226 \tab 1227 \tab 1229 \tab 1231 \tab 1233 \tab 1234 \tab 1235 \tab 1236 \tab 1239 \tab 1313 \tab 1315 \tab 1319 \tab 2131 \tab 2132 \tab 2141 \tab 2142 \tab 2143 \tab 2144 \tab 2145 \tab 2146 \tab 2149 \tab 2221 \tab 2222 \tab 2223 \tab 2230 \tab 2310 \tab 2320 \tab 2331 \tab 2332 \tab 2340 \tab 2359 \tab 2411 \tab 2412 \tab 2419 \tab 2421 \tab 2432 \tab 2441 \tab 2445 \tab 2446 \tab 2451 \tab 2452 \tab 2455 \tab 2460 \tab 3110 \tab 3111 \tab 3112 \tab 3118 \tab 3122 \tab 3131 \tab 3132 \tab 3133 \tab 3141 \tab 3144 \tab 3152 \tab 3211 \tab 3221 \tab 3225 \tab 3226 \tab 3231 \tab 3300 \tab 3340 \tab 3410 \tab 3411 \tab 3412 \tab 3413 \tab 3415 \tab 3416 \tab 3417 \tab 3429 \tab 3432 \tab 3443 \tab 3460 \tab 3471 \tab 4100 \tab 4110 \tab 4111 \tab 4113 \tab 4114 \tab 4115 \tab 4121 \tab 4131 \tab 4132 \tab 4133 \tab 4141 \tab 4142 \tab 4190 \tab 4211 \tab 4212 \tab 4213 \tab 4215 \tab 4221 \tab 4222 \tab 4223 \tab 5111 \tab 5112 \tab 5120 \tab 5122 \tab 5123 \tab 5131 \tab 5141 \tab 5161 \tab 5162 \tab 5169 \tab 5220 \tab 6113 \tab 6129 \tab 6141 \tab 7122 \tab 7123 \tab 7124 \tab 7129 \tab 7131 \tab 7132 \tab 7136 \tab 7137 \tab 7141 \tab 7200 \tab 7212 \tab 7213 \tab 7214 \tab 7215 \tab 7222 \tab 7223 \tab 7230 \tab 7231 \tab 7232 \tab 7240 \tab 7241 \tab 7243 \tab 7324 \tab 7343 \tab 7411 \tab 7412 \tab 7415 \tab 7422 \tab 7433 \tab 7435 \tab 7437 \tab 7442 \tab 7510 \tab 8122 \tab 8141 \tab 8151 \tab 8152 \tab 8161 \tab 8162 \tab 8200 \tab 8221 \tab 8223 \tab 8231 \tab 8232 \tab 8251 \tab 8253 \tab 8261 \tab 8262 \tab 8263 \tab 8264 \tab 8266 \tab 8269 \tab 8282 \tab 8290 \tab 8311 \tab 8312 \tab 8322 \tab 8323 \tab 8332 \tab 8333 \tab 8334 \tab 8340 \tab 9110 \tab 9113 \tab 9131 \tab 9132 \tab 9140 \tab 9142 \tab 9151 \tab 9152 \tab 9153 \tab 9161 \tab 9211 \tab 9212 \tab 9312 \tab 9313 \tab 9321 \tab 9322 \tab 9333 \tab iap \tab uncodeable \tab 1232 \tab 1311 \tab 2113 \tab 2147 \tab 2211 \tab 2453 \tab 2454 \tab 3121 \tab 3142 \tab 3224 \tab 3229 \tab 3434 \tab 3480 \tab 4122 \tab 4143 \tab 5132 \tab 5133 \tab 5163 \tab 5200 \tab 7224 \tab 7233 \tab 7245 \tab 7311 \tab 7341 \tab 8100 \tab 8111 \tab 8113 \tab 8155 \tab 8163 \tab 8224 \tab 8230 \tab 8270 \tab 8275 \tab 9311 \tab 2121 \tab 2148 \tab 2422 \tab 3223 \tab 5143 \tab 7111 \tab 7143 \tab 7300 \tab 8212 \tab 9141 \tab 9320 \tab 1110 \tab 2224 \tab 2431 \tab 2440 \tab 3212 \tab 3470 \tab 3475 \tab 7112 \tab 7134 \tab 7313 \tab 2111 \tab 2122 \tab 2442 \tab 3151 \tab 3450 \tab 3472 \tab 5139 \tab 7133 \tab 8310 \tab 9213 \tab 2213 \tab 3143 \tab 3442 \tab 7420 \tab 2444 \tab 3115 \tab 2110 \tab 3227 \tab 5100 \tab 8123 \tab 5113 \tab 7135 \tab 2229 \tab 3114 \tab 2112 \tab 3400 \tab 7216 \tab 2212 \tab 9160 \tab 3422 \tab 1237 \tab 6124 \tab Total \cr
   1972 \tab 17 \tab 1 \tab 4 \tab 38 \tab 6 \tab 12 \tab 2 \tab 11 \tab 7 \tab 8 \tab 15 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 6 \tab 4 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 10 \tab 1 \tab 3 \tab 3 \tab 23 \tab 11 \tab 19 \tab 31 \tab 6 \tab 1 \tab 7 \tab 8 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 4 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 2 \tab 4 \tab 3 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 16 \tab 3 \tab 2 \tab 4 \tab 2 \tab 24 \tab 6 \tab 1 \tab 1 \tab 2 \tab 5 \tab 7 \tab 13 \tab 6 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 3 \tab 3 \tab 29 \tab 20 \tab 5 \tab 1 \tab 69 \tab 28 \tab 13 \tab 2 \tab 1 \tab 11 \tab 15 \tab 8 \tab 18 \tab 6 \tab 1 \tab 1 \tab 2 \tab 4 \tab 14 \tab 1 \tab 3 \tab 1 \tab 23 \tab 33 \tab 8 \tab 10 \tab 4 \tab 6 \tab 4 \tab 27 \tab 4 \tab 1 \tab 4 \tab 1 \tab 1 \tab 12 \tab 1 \tab 3 \tab 3 \tab 9 \tab 9 \tab 6 \tab 5 \tab 14 \tab 7 \tab 3 \tab 1 \tab 11 \tab 2 \tab 7 \tab 7 \tab 2 \tab 1 \tab 3 \tab 3 \tab 4 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 6 \tab 3 \tab 1 \tab 1 \tab 26 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 5 \tab 6 \tab 1 \tab 1 \tab 3 \tab 3 \tab 9 \tab 1 \tab 8 \tab 1 \tab 19 \tab 9 \tab 2 \tab 3 \tab 3 \tab 2 \tab 3 \tab 2 \tab 45 \tab 1 \tab 8 \tab 4 \tab 11 \tab 2 \tab 3 \tab 1 \tab 53 \tab 1 \tab 25 \tab 1 \tab 6 \tab 2 \tab 1 \tab 2 \tab 12 \tab 1 \tab 11 \tab 2 \tab 1 \tab 11 \tab 18 \tab 167 \tab 167 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 9 \tab 6 \tab 4 \tab 19 \tab 7 \tab 8 \tab 2 \tab 11 \tab 8 \tab 6 \tab 19 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab 2 \tab 9 \tab 3 \tab 3 \tab 2 \tab 7 \tab 1 \tab 1 \tab 5 \tab - \tab 10 \tab 1 \tab 1 \tab - \tab 13 \tab 14 \tab 23 \tab 26 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 2 \tab 5 \tab 3 \tab - \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 20 \tab 4 \tab 3 \tab 4 \tab 3 \tab 25 \tab 4 \tab - \tab 3 \tab - \tab 12 \tab 4 \tab 19 \tab 2 \tab 3 \tab 4 \tab 1 \tab 2 \tab 2 \tab 4 \tab 3 \tab 18 \tab 14 \tab 5 \tab 2 \tab 80 \tab 39 \tab 19 \tab 5 \tab - \tab 4 \tab 11 \tab 1 \tab 23 \tab 4 \tab 1 \tab 1 \tab - \tab 7 \tab 8 \tab - \tab 2 \tab 1 \tab 17 \tab 29 \tab 10 \tab 15 \tab 4 \tab 4 \tab 11 \tab 42 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 20 \tab 3 \tab - \tab - \tab 6 \tab 6 \tab 3 \tab 19 \tab 6 \tab 4 \tab 2 \tab 1 \tab 14 \tab 5 \tab 15 \tab 11 \tab 1 \tab 5 \tab 3 \tab 5 \tab 3 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 8 \tab 2 \tab - \tab - \tab 16 \tab 1 \tab 4 \tab 2 \tab 3 \tab 2 \tab 2 \tab 15 \tab 1 \tab - \tab - \tab 7 \tab 4 \tab - \tab 1 \tab 3 \tab 18 \tab 8 \tab 2 \tab 6 \tab 6 \tab 5 \tab 1 \tab 1 \tab 31 \tab 4 \tab 4 \tab 3 \tab 4 \tab 1 \tab 4 \tab 1 \tab 30 \tab 3 \tab 24 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 18 \tab - \tab 12 \tab 2 \tab 1 \tab 9 \tab 8 \tab 164 \tab 81 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 5 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 18 \tab 3 \tab 6 \tab 22 \tab 5 \tab 7 \tab 2 \tab 12 \tab 11 \tab 6 \tab 27 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 9 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab - \tab 2 \tab - \tab 5 \tab 3 \tab - \tab - \tab 17 \tab 11 \tab 30 \tab 40 \tab 6 \tab 1 \tab 1 \tab 7 \tab 1 \tab 2 \tab 5 \tab 4 \tab - \tab 2 \tab 12 \tab 3 \tab 4 \tab 1 \tab 5 \tab - \tab - \tab 3 \tab 4 \tab 4 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 19 \tab 4 \tab 4 \tab 5 \tab 3 \tab 33 \tab 6 \tab - \tab 3 \tab 1 \tab 11 \tab 3 \tab 13 \tab 5 \tab 2 \tab 1 \tab 5 \tab - \tab - \tab 4 \tab 3 \tab 12 \tab 27 \tab 4 \tab 7 \tab 67 \tab 29 \tab 17 \tab 2 \tab 1 \tab 5 \tab 14 \tab 8 \tab 24 \tab 4 \tab 3 \tab 1 \tab 2 \tab 3 \tab 13 \tab 1 \tab 1 \tab 3 \tab 20 \tab 30 \tab 7 \tab 7 \tab 1 \tab 6 \tab 10 \tab 29 \tab 6 \tab - \tab 2 \tab 5 \tab 1 \tab 19 \tab - \tab 2 \tab 1 \tab 13 \tab 4 \tab 2 \tab 7 \tab 7 \tab 4 \tab 1 \tab - \tab 4 \tab 3 \tab 10 \tab 11 \tab 1 \tab 1 \tab 4 \tab 8 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 5 \tab 1 \tab - \tab - \tab 20 \tab 1 \tab 3 \tab 2 \tab 2 \tab - \tab 2 \tab 15 \tab 2 \tab 2 \tab - \tab 8 \tab 7 \tab 2 \tab 2 \tab 3 \tab 19 \tab 6 \tab 1 \tab 2 \tab 9 \tab 5 \tab 3 \tab 1 \tab 25 \tab 8 \tab 7 \tab 5 \tab 3 \tab 1 \tab 5 \tab - \tab 31 \tab 5 \tab 21 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 8 \tab 1 \tab 11 \tab - \tab 1 \tab 7 \tab 17 \tab 179 \tab 46 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 14 \tab 1 \tab 3 \tab 18 \tab 11 \tab 9 \tab 6 \tab 11 \tab 10 \tab 11 \tab 25 \tab - \tab 3 \tab - \tab 1 \tab 3 \tab 1 \tab 6 \tab 7 \tab 1 \tab 2 \tab 1 \tab 8 \tab 1 \tab 5 \tab 1 \tab 4 \tab - \tab - \tab - \tab 17 \tab 5 \tab 15 \tab 27 \tab 3 \tab - \tab 3 \tab 12 \tab 3 \tab 2 \tab 3 \tab 2 \tab 2 \tab - \tab 4 \tab 3 \tab 3 \tab 1 \tab 3 \tab 2 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 16 \tab 7 \tab 5 \tab 2 \tab - \tab 36 \tab 7 \tab - \tab 2 \tab 1 \tab 7 \tab 3 \tab 11 \tab 5 \tab - \tab 3 \tab 2 \tab - \tab - \tab 6 \tab 1 \tab 13 \tab 18 \tab 3 \tab 5 \tab 75 \tab 36 \tab 23 \tab 1 \tab 1 \tab 5 \tab 16 \tab 1 \tab 19 \tab 5 \tab 3 \tab - \tab 1 \tab 11 \tab 7 \tab - \tab 1 \tab 2 \tab 17 \tab 39 \tab 11 \tab 13 \tab 1 \tab 8 \tab 4 \tab 24 \tab 3 \tab 2 \tab 4 \tab 4 \tab 3 \tab 21 \tab 1 \tab 1 \tab - \tab 7 \tab 9 \tab 6 \tab 11 \tab 8 \tab 2 \tab 2 \tab - \tab 12 \tab 2 \tab 8 \tab 19 \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 8 \tab - \tab 2 \tab 3 \tab 6 \tab 1 \tab - \tab 1 \tab 14 \tab - \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 6 \tab 2 \tab - \tab - \tab 8 \tab 4 \tab - \tab 5 \tab 2 \tab 38 \tab 5 \tab 1 \tab 2 \tab 8 \tab 7 \tab - \tab 2 \tab 29 \tab 5 \tab 4 \tab 5 \tab 3 \tab - \tab 3 \tab - \tab 38 \tab 2 \tab 23 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 11 \tab 3 \tab 11 \tab 3 \tab 2 \tab 8 \tab 11 \tab 140 \tab 106 \tab - \tab - \tab 4 \tab - \tab - \tab 5 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 11 \tab 9 \tab 10 \tab 34 \tab 4 \tab 10 \tab 3 \tab 14 \tab 4 \tab 6 \tab 35 \tab 3 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 10 \tab 4 \tab 2 \tab 3 \tab 3 \tab 3 \tab - \tab 3 \tab 1 \tab 10 \tab 2 \tab - \tab - \tab 16 \tab 10 \tab 22 \tab 28 \tab 2 \tab 1 \tab 4 \tab 7 \tab 3 \tab - \tab 4 \tab 3 \tab 1 \tab 2 \tab 10 \tab 3 \tab 3 \tab 1 \tab 6 \tab - \tab - \tab 3 \tab 4 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 13 \tab 5 \tab 3 \tab 1 \tab 4 \tab 25 \tab 8 \tab - \tab - \tab - \tab 12 \tab 6 \tab 10 \tab 5 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 18 \tab 20 \tab 8 \tab 3 \tab 70 \tab 37 \tab 26 \tab 5 \tab - \tab 9 \tab 7 \tab 4 \tab 17 \tab 1 \tab - \tab - \tab - \tab 8 \tab 17 \tab 3 \tab - \tab 1 \tab 24 \tab 34 \tab 6 \tab 8 \tab 4 \tab 8 \tab 2 \tab 38 \tab 3 \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 5 \tab 4 \tab 5 \tab 18 \tab 12 \tab 2 \tab 1 \tab - \tab 7 \tab 1 \tab 10 \tab 17 \tab 2 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 7 \tab 1 \tab - \tab 23 \tab 6 \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab 21 \tab 1 \tab - \tab - \tab 6 \tab 4 \tab 2 \tab 4 \tab 2 \tab 24 \tab 9 \tab 2 \tab 1 \tab 9 \tab 1 \tab 1 \tab - \tab 30 \tab 3 \tab 6 \tab 2 \tab 1 \tab 1 \tab 8 \tab 2 \tab 36 \tab 3 \tab 19 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 12 \tab 1 \tab 9 \tab 2 \tab 2 \tab 8 \tab 19 \tab 149 \tab 47 \tab 2 \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 3 \tab 2 \tab - \tab - \tab 4 \tab - \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 7 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 18 \tab 4 \tab 3 \tab 24 \tab 5 \tab 10 \tab 3 \tab 10 \tab 5 \tab 6 \tab 19 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 7 \tab 1 \tab 3 \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab 2 \tab 11 \tab 2 \tab 3 \tab - \tab 18 \tab 11 \tab 27 \tab 35 \tab 4 \tab - \tab 5 \tab 10 \tab 1 \tab 5 \tab 2 \tab 5 \tab 2 \tab 1 \tab 6 \tab - \tab 1 \tab 1 \tab 8 \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab 3 \tab - \tab - \tab 13 \tab 4 \tab 1 \tab - \tab 1 \tab 27 \tab 7 \tab - \tab - \tab 1 \tab 5 \tab 10 \tab 5 \tab 1 \tab 2 \tab 2 \tab 4 \tab - \tab - \tab 5 \tab - \tab 19 \tab 11 \tab 3 \tab 7 \tab 72 \tab 27 \tab 17 \tab 5 \tab 2 \tab 4 \tab 9 \tab 8 \tab 22 \tab 6 \tab 1 \tab 1 \tab 3 \tab 5 \tab 10 \tab - \tab - \tab 5 \tab 15 \tab 37 \tab 6 \tab 14 \tab - \tab 2 \tab 6 \tab 30 \tab 4 \tab - \tab 3 \tab 4 \tab - \tab 22 \tab 1 \tab - \tab 1 \tab 4 \tab 7 \tab 10 \tab 13 \tab 15 \tab 3 \tab 4 \tab - \tab 7 \tab 2 \tab 9 \tab 21 \tab - \tab 3 \tab - \tab 5 \tab 6 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 5 \tab - \tab 3 \tab - \tab 15 \tab 1 \tab 1 \tab 5 \tab 2 \tab - \tab 2 \tab 16 \tab 2 \tab - \tab 1 \tab 4 \tab 4 \tab - \tab 1 \tab 4 \tab 24 \tab 8 \tab 2 \tab 5 \tab 1 \tab 9 \tab 1 \tab - \tab 33 \tab 3 \tab 8 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 31 \tab 1 \tab 20 \tab 3 \tab 2 \tab 4 \tab 2 \tab 3 \tab 13 \tab - \tab 15 \tab 4 \tab 2 \tab 15 \tab 20 \tab 116 \tab 154 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 4 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 6 \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 22 \tab 5 \tab 6 \tab 15 \tab 6 \tab 13 \tab 4 \tab 17 \tab 8 \tab 2 \tab 35 \tab 3 \tab 4 \tab - \tab 2 \tab 2 \tab 1 \tab 8 \tab 5 \tab - \tab 5 \tab 4 \tab - \tab - \tab 4 \tab 3 \tab 7 \tab 5 \tab 2 \tab - \tab 20 \tab 7 \tab 22 \tab 32 \tab 1 \tab 6 \tab 7 \tab 9 \tab 3 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab 6 \tab 2 \tab 3 \tab 2 \tab 3 \tab 2 \tab 3 \tab 2 \tab 3 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 22 \tab 9 \tab 5 \tab 5 \tab 3 \tab 27 \tab 3 \tab - \tab - \tab 3 \tab 10 \tab 8 \tab 18 \tab 3 \tab 3 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 13 \tab 17 \tab 5 \tab 5 \tab 87 \tab 46 \tab 18 \tab 2 \tab 2 \tab 4 \tab 4 \tab 5 \tab 15 \tab 7 \tab - \tab 1 \tab - \tab 7 \tab 11 \tab - \tab 1 \tab 2 \tab 22 \tab 44 \tab 6 \tab 17 \tab 5 \tab 7 \tab 6 \tab 43 \tab 11 \tab - \tab 3 \tab - \tab - \tab 21 \tab 5 \tab 3 \tab 2 \tab 2 \tab 10 \tab 3 \tab 20 \tab 10 \tab 1 \tab 1 \tab - \tab 13 \tab 4 \tab 15 \tab 12 \tab - \tab 2 \tab - \tab 7 \tab 4 \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 8 \tab 1 \tab - \tab 1 \tab 20 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 4 \tab 16 \tab 2 \tab - \tab - \tab 8 \tab 9 \tab - \tab 5 \tab 1 \tab 31 \tab 10 \tab 1 \tab - \tab 9 \tab 14 \tab 1 \tab 1 \tab 33 \tab 5 \tab 5 \tab 5 \tab 3 \tab 1 \tab 5 \tab 1 \tab 30 \tab 3 \tab 18 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 9 \tab - \tab 6 \tab 1 \tab 2 \tab 7 \tab 19 \tab 131 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 13 \tab 6 \tab 6 \tab 25 \tab 8 \tab 15 \tab 5 \tab 19 \tab 10 \tab 7 \tab 36 \tab 3 \tab - \tab - \tab 1 \tab 6 \tab 2 \tab 23 \tab 12 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 2 \tab 1 \tab 10 \tab 3 \tab - \tab 1 \tab 19 \tab 8 \tab 17 \tab 32 \tab 3 \tab - \tab - \tab 13 \tab 3 \tab 4 \tab 8 \tab 4 \tab - \tab - \tab 9 \tab 3 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 5 \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 12 \tab 7 \tab - \tab 3 \tab 1 \tab 29 \tab 4 \tab - \tab 1 \tab 4 \tab 8 \tab 9 \tab 11 \tab 7 \tab 2 \tab 3 \tab 8 \tab - \tab - \tab 4 \tab 3 \tab 12 \tab 14 \tab 8 \tab 4 \tab 77 \tab 33 \tab 12 \tab 3 \tab 6 \tab 3 \tab 9 \tab 4 \tab 12 \tab 4 \tab - \tab - \tab 1 \tab 10 \tab 7 \tab - \tab - \tab 3 \tab 21 \tab 39 \tab 10 \tab 8 \tab 4 \tab 5 \tab 8 \tab 35 \tab 5 \tab 1 \tab 1 \tab 5 \tab 1 \tab 14 \tab 4 \tab 1 \tab - \tab 2 \tab 5 \tab 7 \tab 11 \tab 12 \tab 2 \tab 1 \tab - \tab 14 \tab 2 \tab 12 \tab 20 \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 8 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 17 \tab 4 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 13 \tab 1 \tab 1 \tab - \tab 13 \tab 8 \tab 1 \tab - \tab 2 \tab 24 \tab 1 \tab 4 \tab 3 \tab 6 \tab 5 \tab 3 \tab - \tab 32 \tab 2 \tab 6 \tab 3 \tab 1 \tab 1 \tab 4 \tab - \tab 28 \tab 2 \tab 22 \tab - \tab 3 \tab 2 \tab - \tab - \tab 11 \tab - \tab 7 \tab 1 \tab - \tab 2 \tab 13 \tab 122 \tab 45 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 1 \tab - \tab 6 \tab 1 \tab 3 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 11 \tab 7 \tab 6 \tab 22 \tab 14 \tab 13 \tab 1 \tab 15 \tab 8 \tab 6 \tab 42 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 18 \tab 2 \tab 3 \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 10 \tab 2 \tab 1 \tab - \tab 27 \tab 9 \tab 19 \tab 44 \tab 5 \tab 6 \tab 7 \tab 8 \tab 2 \tab - \tab 10 \tab 3 \tab 1 \tab - \tab 13 \tab 1 \tab 4 \tab - \tab 4 \tab 2 \tab 2 \tab 1 \tab 3 \tab 7 \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 17 \tab 7 \tab 7 \tab 3 \tab 8 \tab 50 \tab 7 \tab 1 \tab 2 \tab 1 \tab 8 \tab 4 \tab 12 \tab 4 \tab 5 \tab 5 \tab 4 \tab 1 \tab - \tab 7 \tab 2 \tab 22 \tab 16 \tab 6 \tab 4 \tab 76 \tab 32 \tab 25 \tab 13 \tab 7 \tab 8 \tab 17 \tab 5 \tab 30 \tab 8 \tab 1 \tab 2 \tab 1 \tab 8 \tab 11 \tab - \tab 1 \tab 2 \tab 25 \tab 40 \tab 17 \tab 25 \tab 2 \tab 6 \tab 10 \tab 43 \tab 6 \tab 1 \tab 3 \tab 3 \tab 2 \tab 13 \tab 2 \tab - \tab 2 \tab 7 \tab 9 \tab 6 \tab 23 \tab 13 \tab 5 \tab 4 \tab 1 \tab 12 \tab 4 \tab 22 \tab 22 \tab 4 \tab 7 \tab 4 \tab 9 \tab 6 \tab - \tab 10 \tab 3 \tab 4 \tab 1 \tab 5 \tab 4 \tab 2 \tab 1 \tab 23 \tab - \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 22 \tab 1 \tab - \tab - \tab 11 \tab 9 \tab 1 \tab 1 \tab 7 \tab 22 \tab 8 \tab - \tab 3 \tab 10 \tab 5 \tab 2 \tab 1 \tab 34 \tab 4 \tab 11 \tab 4 \tab 3 \tab - \tab 2 \tab 1 \tab 75 \tab 6 \tab 26 \tab - \tab 1 \tab - \tab 4 \tab 2 \tab 16 \tab 2 \tab 21 \tab 1 \tab 3 \tab 8 \tab 19 \tab 136 \tab 32 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 5 \tab - \tab - \tab 2 \tab 3 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 7 \tab 2 \tab - \tab - \tab 4 \tab 2 \tab 3 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 8 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 12 \tab 8 \tab 10 \tab 30 \tab 4 \tab 17 \tab 2 \tab 23 \tab 7 \tab 6 \tab 41 \tab 1 \tab 2 \tab 4 \tab - \tab 2 \tab 2 \tab 13 \tab 7 \tab 3 \tab 4 \tab 1 \tab 4 \tab - \tab 2 \tab 1 \tab 11 \tab 2 \tab 2 \tab 2 \tab 21 \tab 19 \tab 21 \tab 34 \tab 3 \tab 1 \tab 2 \tab 18 \tab 2 \tab 5 \tab 5 \tab 8 \tab - \tab - \tab 14 \tab 6 \tab 1 \tab 2 \tab 4 \tab 3 \tab 2 \tab - \tab 2 \tab 7 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab 18 \tab 8 \tab 7 \tab 9 \tab 6 \tab 28 \tab 6 \tab - \tab 3 \tab 3 \tab 6 \tab 7 \tab 22 \tab 2 \tab 3 \tab 3 \tab 7 \tab - \tab - \tab 4 \tab 1 \tab 15 \tab 13 \tab 10 \tab 6 \tab 98 \tab 45 \tab 19 \tab 6 \tab - \tab 8 \tab 5 \tab 2 \tab 30 \tab 3 \tab - \tab 1 \tab - \tab 9 \tab 5 \tab - \tab - \tab 3 \tab 34 \tab 42 \tab 9 \tab 13 \tab 1 \tab 4 \tab 7 \tab 36 \tab 8 \tab - \tab 1 \tab 6 \tab 3 \tab 15 \tab 1 \tab 2 \tab - \tab 8 \tab 11 \tab 5 \tab 20 \tab 12 \tab 4 \tab 3 \tab - \tab 14 \tab 3 \tab 11 \tab 14 \tab 3 \tab 3 \tab 2 \tab 6 \tab - \tab - \tab 5 \tab 2 \tab 1 \tab 5 \tab 4 \tab - \tab 2 \tab 1 \tab 9 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 16 \tab 3 \tab - \tab 1 \tab 9 \tab 1 \tab 1 \tab 3 \tab 2 \tab 22 \tab 5 \tab - \tab 2 \tab 9 \tab 8 \tab 1 \tab 2 \tab 23 \tab 3 \tab 5 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 28 \tab 2 \tab 15 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 10 \tab - \tab 11 \tab 2 \tab 1 \tab 8 \tab 20 \tab 106 \tab 27 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 3 \tab 2 \tab 4 \tab 1 \tab - \tab 5 \tab 2 \tab 2 \tab 3 \tab 1 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 17 \tab 9 \tab 7 \tab 21 \tab 5 \tab 19 \tab - \tab 22 \tab 6 \tab 8 \tab 43 \tab 1 \tab 4 \tab - \tab 2 \tab 4 \tab 1 \tab 20 \tab 9 \tab 1 \tab 3 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab 13 \tab 5 \tab - \tab - \tab 25 \tab 8 \tab 25 \tab 18 \tab 7 \tab 2 \tab 4 \tab 15 \tab 6 \tab 5 \tab 5 \tab 3 \tab 1 \tab 1 \tab 11 \tab 4 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 5 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 17 \tab 5 \tab 3 \tab 4 \tab 4 \tab 29 \tab 9 \tab 1 \tab 5 \tab 2 \tab 5 \tab 6 \tab 20 \tab 6 \tab 2 \tab 6 \tab 9 \tab - \tab - \tab 3 \tab 1 \tab 13 \tab 12 \tab 5 \tab 2 \tab 65 \tab 32 \tab 24 \tab 5 \tab 3 \tab 1 \tab 5 \tab 9 \tab 22 \tab 5 \tab 1 \tab - \tab 3 \tab 7 \tab 8 \tab - \tab 1 \tab 3 \tab 22 \tab 38 \tab 12 \tab 17 \tab 3 \tab 8 \tab 4 \tab 55 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 18 \tab 5 \tab 1 \tab - \tab 6 \tab 4 \tab 3 \tab 18 \tab 6 \tab 3 \tab 2 \tab - \tab 10 \tab 6 \tab 10 \tab 14 \tab 2 \tab 2 \tab 1 \tab 5 \tab 5 \tab - \tab 5 \tab 3 \tab 3 \tab 3 \tab 4 \tab 3 \tab - \tab - \tab 13 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab 13 \tab 4 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 3 \tab - \tab 15 \tab 12 \tab 5 \tab 3 \tab 6 \tab 3 \tab 1 \tab - \tab 26 \tab 3 \tab 7 \tab 2 \tab 7 \tab 1 \tab 6 \tab - \tab 30 \tab 3 \tab 26 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 10 \tab - \tab 11 \tab 1 \tab 1 \tab 9 \tab 12 \tab 94 \tab 1 \tab 5 \tab 1 \tab 4 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 4 \tab 8 \tab 8 \tab 23 \tab 5 \tab 13 \tab 7 \tab 24 \tab 13 \tab 13 \tab 45 \tab - \tab 1 \tab 1 \tab 2 \tab 11 \tab 1 \tab 24 \tab 7 \tab 2 \tab 5 \tab 6 \tab 4 \tab 1 \tab 2 \tab 4 \tab 6 \tab 6 \tab 6 \tab - \tab 22 \tab 9 \tab 19 \tab 26 \tab 2 \tab 3 \tab 4 \tab 12 \tab 2 \tab 4 \tab 8 \tab 4 \tab - \tab 3 \tab 7 \tab 3 \tab 5 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 11 \tab 2 \tab - \tab 1 \tab - \tab - \tab 9 \tab 7 \tab 4 \tab 2 \tab 5 \tab 23 \tab 5 \tab - \tab 1 \tab 1 \tab 8 \tab 9 \tab 22 \tab 2 \tab 5 \tab 5 \tab 9 \tab 1 \tab 1 \tab 5 \tab 2 \tab 20 \tab 17 \tab 7 \tab 3 \tab 77 \tab 36 \tab 21 \tab 1 \tab 4 \tab 6 \tab 8 \tab 8 \tab 15 \tab 8 \tab - \tab 2 \tab 1 \tab 12 \tab 8 \tab - \tab 4 \tab 1 \tab 21 \tab 47 \tab 16 \tab 16 \tab 2 \tab 8 \tab 7 \tab 32 \tab 9 \tab - \tab - \tab 5 \tab 2 \tab 13 \tab 1 \tab - \tab 1 \tab 5 \tab 16 \tab 5 \tab 10 \tab 9 \tab 1 \tab 4 \tab 1 \tab 9 \tab 4 \tab 22 \tab 20 \tab 3 \tab 3 \tab 1 \tab 6 \tab 1 \tab - \tab 7 \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab 17 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 7 \tab 1 \tab - \tab - \tab 7 \tab 4 \tab 2 \tab 1 \tab 3 \tab 16 \tab 8 \tab 1 \tab - \tab 4 \tab 3 \tab 2 \tab 1 \tab 33 \tab 1 \tab 7 \tab 1 \tab 4 \tab - \tab 2 \tab 1 \tab 25 \tab 2 \tab 20 \tab 2 \tab 2 \tab 3 \tab - \tab - \tab 16 \tab - \tab 11 \tab 2 \tab - \tab 7 \tab 14 \tab 84 \tab 36 \tab 1 \tab 3 \tab - \tab 2 \tab 4 \tab 5 \tab - \tab - \tab - \tab 1 \tab 4 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 7 \tab 6 \tab 9 \tab 20 \tab 14 \tab 19 \tab 2 \tab 14 \tab 8 \tab 8 \tab 47 \tab 1 \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab 26 \tab 6 \tab 3 \tab 5 \tab 3 \tab 6 \tab 1 \tab 4 \tab - \tab 8 \tab 7 \tab 2 \tab - \tab 30 \tab 18 \tab 17 \tab 27 \tab 3 \tab - \tab 4 \tab 11 \tab 1 \tab 5 \tab 4 \tab 4 \tab - \tab 1 \tab 10 \tab 7 \tab 4 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 6 \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 15 \tab 4 \tab 3 \tab 7 \tab 8 \tab 35 \tab 4 \tab - \tab 1 \tab 3 \tab 7 \tab 7 \tab 15 \tab 7 \tab 1 \tab 3 \tab 7 \tab 2 \tab - \tab 4 \tab 3 \tab 14 \tab 13 \tab 4 \tab 4 \tab 65 \tab 25 \tab 11 \tab 2 \tab - \tab 4 \tab 7 \tab 5 \tab 20 \tab 7 \tab 2 \tab - \tab - \tab 13 \tab 10 \tab 1 \tab - \tab 4 \tab 18 \tab 39 \tab 11 \tab 20 \tab 1 \tab 8 \tab 10 \tab 48 \tab 7 \tab 1 \tab 1 \tab 4 \tab 1 \tab 19 \tab 2 \tab - \tab - \tab 6 \tab 3 \tab 5 \tab 16 \tab 7 \tab 1 \tab 3 \tab - \tab 15 \tab - \tab 6 \tab 13 \tab 1 \tab 4 \tab 2 \tab 5 \tab 5 \tab - \tab 2 \tab 6 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 15 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 4 \tab 20 \tab 2 \tab - \tab 1 \tab 8 \tab 6 \tab - \tab 2 \tab 1 \tab 13 \tab 12 \tab - \tab 5 \tab 10 \tab 5 \tab 1 \tab - \tab 22 \tab 6 \tab 2 \tab 2 \tab 3 \tab - \tab 7 \tab 2 \tab 23 \tab - \tab 22 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 9 \tab - \tab 9 \tab 1 \tab - \tab 9 \tab 17 \tab 104 \tab 8 \tab - \tab - \tab 1 \tab - \tab 3 \tab 5 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 9 \tab 8 \tab 10 \tab 28 \tab 9 \tab 14 \tab 4 \tab 16 \tab 11 \tab 5 \tab 54 \tab 1 \tab 5 \tab 1 \tab 4 \tab 7 \tab 3 \tab 27 \tab 12 \tab 3 \tab 5 \tab 6 \tab 4 \tab 1 \tab 5 \tab 1 \tab 10 \tab 4 \tab 1 \tab - \tab 23 \tab 13 \tab 26 \tab 25 \tab 5 \tab 2 \tab 5 \tab 13 \tab 4 \tab 4 \tab 5 \tab 3 \tab - \tab 1 \tab 12 \tab 4 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab 3 \tab 3 \tab 5 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 15 \tab 7 \tab 9 \tab 3 \tab 5 \tab 65 \tab 10 \tab 3 \tab 3 \tab 2 \tab 7 \tab 7 \tab 17 \tab 3 \tab 6 \tab 6 \tab 7 \tab - \tab - \tab 9 \tab - \tab 15 \tab 11 \tab 10 \tab 6 \tab 74 \tab 29 \tab 25 \tab 8 \tab 3 \tab 9 \tab 12 \tab 13 \tab 32 \tab 5 \tab 3 \tab 3 \tab - \tab 9 \tab 10 \tab 4 \tab - \tab 5 \tab 32 \tab 27 \tab 14 \tab 25 \tab 3 \tab 6 \tab 15 \tab 45 \tab 6 \tab 1 \tab 1 \tab 4 \tab - \tab 23 \tab 2 \tab 4 \tab 2 \tab 4 \tab 8 \tab 5 \tab 12 \tab 10 \tab 4 \tab 2 \tab 1 \tab 7 \tab 2 \tab 15 \tab 13 \tab 3 \tab 2 \tab 3 \tab 4 \tab 3 \tab - \tab 1 \tab 3 \tab 1 \tab 3 \tab 6 \tab 2 \tab 3 \tab - \tab 21 \tab 2 \tab - \tab 3 \tab - \tab - \tab 3 \tab 12 \tab - \tab - \tab - \tab 3 \tab 11 \tab 2 \tab 3 \tab 2 \tab 17 \tab 7 \tab 4 \tab 2 \tab 5 \tab 9 \tab - \tab - \tab 31 \tab 3 \tab 6 \tab 5 \tab 2 \tab - \tab 2 \tab - \tab 61 \tab 2 \tab 29 \tab 2 \tab 3 \tab 5 \tab - \tab - \tab 12 \tab - \tab 16 \tab 4 \tab 3 \tab 10 \tab 24 \tab 105 \tab 72 \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab 7 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 4 \tab 2 \tab 3 \tab - \tab 3 \tab 5 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 7 \tab 5 \tab 10 \tab 10 \tab 8 \tab 12 \tab 4 \tab 20 \tab 7 \tab 9 \tab 57 \tab 1 \tab 2 \tab 3 \tab 5 \tab 2 \tab - \tab 16 \tab 14 \tab 3 \tab 3 \tab 5 \tab 5 \tab 2 \tab 3 \tab - \tab 6 \tab 6 \tab 2 \tab - \tab 24 \tab 13 \tab 11 \tab 25 \tab 6 \tab 1 \tab 5 \tab 15 \tab 7 \tab 3 \tab 4 \tab 3 \tab - \tab 1 \tab 14 \tab 5 \tab 4 \tab 2 \tab 4 \tab 2 \tab 3 \tab 5 \tab 7 \tab 2 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab 9 \tab 9 \tab 3 \tab 2 \tab 3 \tab 33 \tab 5 \tab - \tab 2 \tab 2 \tab 4 \tab 6 \tab 27 \tab 3 \tab 6 \tab 1 \tab 6 \tab - \tab - \tab 5 \tab 2 \tab 14 \tab 19 \tab 6 \tab 5 \tab 67 \tab 33 \tab 22 \tab 4 \tab 3 \tab 5 \tab 10 \tab 3 \tab 19 \tab 5 \tab 2 \tab 1 \tab 1 \tab 8 \tab 9 \tab 1 \tab 3 \tab 4 \tab 27 \tab 29 \tab 16 \tab 12 \tab 4 \tab 6 \tab 4 \tab 57 \tab 8 \tab 1 \tab 2 \tab 2 \tab 2 \tab 18 \tab 3 \tab 3 \tab - \tab 1 \tab 7 \tab 6 \tab 19 \tab 3 \tab 2 \tab 2 \tab - \tab 8 \tab 1 \tab 12 \tab 22 \tab 3 \tab 2 \tab 2 \tab 7 \tab 2 \tab - \tab 8 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 14 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 9 \tab - \tab 1 \tab - \tab 5 \tab 5 \tab - \tab 4 \tab 1 \tab 24 \tab 5 \tab 1 \tab 4 \tab 4 \tab 3 \tab 1 \tab 1 \tab 22 \tab 10 \tab 5 \tab 4 \tab 4 \tab - \tab 1 \tab 1 \tab 35 \tab 1 \tab 20 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 7 \tab - \tab 15 \tab - \tab 2 \tab 6 \tab 15 \tab 88 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 4 \tab 1 \tab 4 \tab 3 \tab 4 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 18 \tab 8 \tab 14 \tab 15 \tab 9 \tab 17 \tab 4 \tab 24 \tab 13 \tab 5 \tab 54 \tab 1 \tab 3 \tab 5 \tab 2 \tab 5 \tab - \tab 30 \tab 7 \tab 2 \tab 5 \tab 4 \tab 5 \tab 1 \tab 4 \tab - \tab 8 \tab 4 \tab 1 \tab - \tab 24 \tab 17 \tab 23 \tab 19 \tab 7 \tab 6 \tab 5 \tab 8 \tab 4 \tab 7 \tab 12 \tab 4 \tab - \tab - \tab 16 \tab 4 \tab 2 \tab 4 \tab 3 \tab 5 \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 4 \tab - \tab - \tab 17 \tab 8 \tab 2 \tab 4 \tab 3 \tab 39 \tab 4 \tab - \tab 2 \tab 4 \tab 7 \tab 5 \tab 22 \tab 3 \tab 5 \tab 4 \tab 10 \tab 1 \tab - \tab 9 \tab 3 \tab 21 \tab 11 \tab 5 \tab 4 \tab 83 \tab 23 \tab 22 \tab 2 \tab - \tab 4 \tab 6 \tab 5 \tab 23 \tab 5 \tab - \tab 1 \tab 2 \tab 11 \tab 7 \tab - \tab - \tab 5 \tab 23 \tab 32 \tab 9 \tab 9 \tab 2 \tab 5 \tab 8 \tab 52 \tab 8 \tab 1 \tab 1 \tab 1 \tab 1 \tab 16 \tab 5 \tab 1 \tab 2 \tab 6 \tab 10 \tab 5 \tab 15 \tab 13 \tab 3 \tab 3 \tab 1 \tab 9 \tab 2 \tab 10 \tab 19 \tab 5 \tab 1 \tab 2 \tab 4 \tab 2 \tab - \tab 5 \tab 2 \tab 2 \tab 3 \tab 5 \tab 3 \tab 1 \tab - \tab 10 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 17 \tab 1 \tab - \tab - \tab 4 \tab 9 \tab 1 \tab 3 \tab - \tab 11 \tab 10 \tab 3 \tab - \tab 5 \tab 2 \tab 1 \tab - \tab 29 \tab 5 \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 24 \tab 1 \tab 25 \tab 2 \tab 5 \tab - \tab - \tab - \tab 9 \tab - \tab 8 \tab 1 \tab 1 \tab 6 \tab 13 \tab 96 \tab 13 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 4 \tab - \tab - \tab 5 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 8 \tab 4 \tab 12 \tab 15 \tab 15 \tab 15 \tab 1 \tab 19 \tab 9 \tab 6 \tab 49 \tab 4 \tab 2 \tab 1 \tab 2 \tab 6 \tab - \tab 20 \tab 4 \tab 2 \tab 5 \tab 2 \tab 10 \tab - \tab 2 \tab - \tab 14 \tab 4 \tab 1 \tab - \tab 21 \tab 13 \tab 20 \tab 23 \tab - \tab 3 \tab 8 \tab 16 \tab 4 \tab 6 \tab 5 \tab 4 \tab - \tab 3 \tab 17 \tab 7 \tab 4 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 13 \tab 6 \tab 3 \tab 2 \tab 5 \tab 29 \tab 8 \tab - \tab 3 \tab 1 \tab 6 \tab 13 \tab 19 \tab 2 \tab 3 \tab 5 \tab 6 \tab - \tab 1 \tab 9 \tab - \tab 17 \tab 10 \tab 5 \tab 2 \tab 75 \tab 27 \tab 19 \tab 4 \tab 1 \tab 8 \tab 12 \tab 6 \tab 16 \tab 3 \tab 4 \tab 1 \tab 1 \tab 10 \tab 5 \tab - \tab - \tab 4 \tab 24 \tab 31 \tab 12 \tab 11 \tab 2 \tab 4 \tab 3 \tab 34 \tab 4 \tab 2 \tab 3 \tab 6 \tab - \tab 10 \tab - \tab - \tab 2 \tab 10 \tab 5 \tab 4 \tab 10 \tab 5 \tab 3 \tab 1 \tab 1 \tab 8 \tab 3 \tab 15 \tab 16 \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 2 \tab 5 \tab 2 \tab 1 \tab - \tab 8 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 16 \tab - \tab - \tab - \tab 2 \tab 4 \tab 2 \tab 4 \tab 2 \tab 13 \tab 4 \tab 2 \tab 1 \tab 5 \tab 4 \tab - \tab - \tab 30 \tab 6 \tab 5 \tab 1 \tab 6 \tab - \tab 1 \tab 1 \tab 20 \tab 3 \tab 11 \tab 1 \tab 5 \tab 2 \tab 1 \tab - \tab 11 \tab - \tab 14 \tab - \tab - \tab 4 \tab 10 \tab 73 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 2 \tab 4 \tab 1 \tab 1 \tab 3 \tab 4 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 12 \tab 6 \tab 5 \tab 10 \tab 7 \tab 12 \tab 3 \tab 16 \tab 6 \tab 8 \tab 50 \tab - \tab 7 \tab 4 \tab 4 \tab 6 \tab 1 \tab 29 \tab 8 \tab 1 \tab 5 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 5 \tab 4 \tab - \tab 1 \tab 18 \tab 12 \tab 20 \tab 30 \tab 8 \tab 5 \tab 4 \tab 13 \tab 7 \tab 4 \tab 6 \tab 2 \tab 1 \tab 1 \tab 12 \tab 9 \tab 3 \tab 3 \tab - \tab 5 \tab 2 \tab 3 \tab 2 \tab 4 \tab - \tab 3 \tab 4 \tab - \tab - \tab 13 \tab 10 \tab 5 \tab 3 \tab 3 \tab 47 \tab 7 \tab - \tab 1 \tab 2 \tab 7 \tab 8 \tab 20 \tab 8 \tab 5 \tab 4 \tab 10 \tab - \tab 1 \tab 8 \tab 2 \tab 18 \tab 14 \tab 15 \tab 4 \tab 66 \tab 29 \tab 18 \tab 2 \tab 5 \tab 5 \tab 10 \tab 7 \tab 24 \tab 5 \tab 1 \tab 1 \tab 1 \tab 11 \tab 5 \tab - \tab - \tab 4 \tab 28 \tab 35 \tab 12 \tab 14 \tab 2 \tab 6 \tab 10 \tab 38 \tab 7 \tab 1 \tab - \tab 2 \tab 1 \tab 15 \tab 3 \tab 1 \tab 2 \tab 6 \tab 6 \tab 1 \tab 15 \tab 8 \tab 1 \tab 1 \tab 1 \tab 9 \tab 3 \tab 10 \tab 12 \tab - \tab 1 \tab 1 \tab 10 \tab 5 \tab - \tab 3 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 13 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 25 \tab 1 \tab - \tab - \tab 3 \tab 8 \tab 1 \tab - \tab 3 \tab 27 \tab 8 \tab 2 \tab 3 \tab 4 \tab 6 \tab 1 \tab 1 \tab 24 \tab 4 \tab 7 \tab 3 \tab - \tab - \tab - \tab 4 \tab 12 \tab 1 \tab 21 \tab 2 \tab 4 \tab 2 \tab - \tab 2 \tab 12 \tab - \tab 8 \tab - \tab 1 \tab 2 \tab 16 \tab 99 \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab 4 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab 8 \tab - \tab - \tab 5 \tab 2 \tab 2 \tab 1 \tab 4 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 3 \tab 3 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab - \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 9 \tab 10 \tab 9 \tab 20 \tab 8 \tab 11 \tab 5 \tab 25 \tab 18 \tab 5 \tab 59 \tab 1 \tab 6 \tab 1 \tab 7 \tab 9 \tab 2 \tab 29 \tab 5 \tab 7 \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 14 \tab 7 \tab 1 \tab 1 \tab 23 \tab 7 \tab 15 \tab 25 \tab 3 \tab 5 \tab 6 \tab 21 \tab 6 \tab 3 \tab 15 \tab 2 \tab - \tab 3 \tab 18 \tab 5 \tab 6 \tab 3 \tab 3 \tab 2 \tab 2 \tab 3 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 17 \tab 10 \tab 6 \tab 6 \tab 12 \tab 30 \tab 8 \tab 2 \tab 5 \tab 5 \tab 9 \tab 12 \tab 21 \tab 8 \tab 2 \tab 4 \tab 16 \tab - \tab - \tab 7 \tab 1 \tab 13 \tab 16 \tab 7 \tab 4 \tab 73 \tab 29 \tab 21 \tab 1 \tab 1 \tab 6 \tab 6 \tab 11 \tab 24 \tab 6 \tab 1 \tab 2 \tab 4 \tab 7 \tab 9 \tab 2 \tab - \tab 2 \tab 22 \tab 34 \tab 12 \tab 23 \tab 3 \tab 6 \tab 7 \tab 40 \tab 6 \tab 1 \tab - \tab - \tab - \tab 13 \tab 3 \tab 3 \tab - \tab 8 \tab 12 \tab 3 \tab 16 \tab 6 \tab 3 \tab 2 \tab - \tab 6 \tab 2 \tab 16 \tab 18 \tab 2 \tab 1 \tab - \tab 9 \tab 6 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 4 \tab 3 \tab - \tab - \tab 12 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 2 \tab 18 \tab 2 \tab - \tab - \tab 6 \tab 7 \tab 3 \tab 4 \tab - \tab 20 \tab 3 \tab 1 \tab 2 \tab 4 \tab 3 \tab - \tab 1 \tab 31 \tab 5 \tab 5 \tab 1 \tab 7 \tab - \tab 4 \tab - \tab 25 \tab 2 \tab 20 \tab 1 \tab 4 \tab 3 \tab 1 \tab - \tab 7 \tab - \tab 12 \tab - \tab 2 \tab 5 \tab 16 \tab 80 \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 5 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab 6 \tab - \tab 1 \tab 3 \tab 3 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1606 \cr
   1994 \tab 24 \tab 26 \tab 19 \tab 22 \tab 16 \tab 26 \tab 3 \tab 31 \tab 22 \tab 18 \tab 87 \tab 8 \tab 7 \tab 2 \tab 8 \tab 12 \tab 2 \tab 57 \tab 25 \tab 9 \tab 8 \tab 8 \tab 5 \tab - \tab 9 \tab 4 \tab 23 \tab 9 \tab 1 \tab 2 \tab 39 \tab 25 \tab 33 \tab 61 \tab 4 \tab 3 \tab 11 \tab 37 \tab 14 \tab 9 \tab 14 \tab 4 \tab 1 \tab 5 \tab 32 \tab 12 \tab 10 \tab 4 \tab 6 \tab 5 \tab 5 \tab 4 \tab 5 \tab 9 \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 28 \tab 11 \tab 16 \tab 4 \tab 16 \tab 76 \tab 21 \tab 3 \tab 8 \tab 3 \tab 10 \tab 15 \tab 53 \tab 10 \tab 5 \tab 11 \tab 32 \tab - \tab 1 \tab 11 \tab 1 \tab 18 \tab 29 \tab 12 \tab 8 \tab 143 \tab 53 \tab 43 \tab 3 \tab 2 \tab 12 \tab 26 \tab 15 \tab 44 \tab 6 \tab 3 \tab - \tab 2 \tab 18 \tab 14 \tab 1 \tab 2 \tab 4 \tab 31 \tab 77 \tab 33 \tab 33 \tab 7 \tab 19 \tab 14 \tab 88 \tab 16 \tab 3 \tab 2 \tab 4 \tab 3 \tab 23 \tab 4 \tab 4 \tab 3 \tab 17 \tab 20 \tab 7 \tab 25 \tab 15 \tab 2 \tab 5 \tab 1 \tab 6 \tab 5 \tab 23 \tab 36 \tab 4 \tab 4 \tab 2 \tab 9 \tab 2 \tab 1 \tab 6 \tab 12 \tab 1 \tab 1 \tab 4 \tab 3 \tab 2 \tab 1 \tab 24 \tab 1 \tab 8 \tab 1 \tab 3 \tab 1 \tab 4 \tab 36 \tab 2 \tab - \tab 1 \tab 6 \tab 11 \tab 2 \tab 4 \tab 2 \tab 17 \tab 9 \tab 4 \tab 2 \tab 16 \tab 10 \tab 1 \tab - \tab 63 \tab 13 \tab 12 \tab 2 \tab 8 \tab - \tab 5 \tab 2 \tab 48 \tab 5 \tab 37 \tab 6 \tab 5 \tab 4 \tab 2 \tab 1 \tab 13 \tab - \tab 20 \tab 2 \tab 3 \tab 8 \tab 33 \tab 138 \tab 9 \tab 7 \tab 2 \tab 2 \tab - \tab 3 \tab 7 \tab 3 \tab 4 \tab - \tab 1 \tab 13 \tab - \tab - \tab 8 \tab 1 \tab 1 \tab 9 \tab 13 \tab 4 \tab - \tab - \tab 9 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 3 \tab 7 \tab - \tab 2 \tab 4 \tab 2 \tab 3 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 6 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 4 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2992 \cr
   1996 \tab 4 \tab 19 \tab 28 \tab 15 \tab 18 \tab 37 \tab 5 \tab 35 \tab 21 \tab 9 \tab 94 \tab 3 \tab 11 \tab 3 \tab 7 \tab 15 \tab - \tab 52 \tab 30 \tab 7 \tab 9 \tab 9 \tab 10 \tab 4 \tab 8 \tab - \tab 27 \tab 16 \tab 4 \tab - \tab 42 \tab 15 \tab 36 \tab 50 \tab 13 \tab 9 \tab 12 \tab 31 \tab 16 \tab 14 \tab 9 \tab 10 \tab - \tab 3 \tab 34 \tab 6 \tab 7 \tab 5 \tab 9 \tab 3 \tab 3 \tab 3 \tab 4 \tab 4 \tab 4 \tab - \tab 6 \tab - \tab - \tab 21 \tab 9 \tab 9 \tab 2 \tab 14 \tab 74 \tab 17 \tab 1 \tab 4 \tab 4 \tab 8 \tab 11 \tab 43 \tab 9 \tab 8 \tab 12 \tab 25 \tab 3 \tab - \tab 17 \tab - \tab 33 \tab 21 \tab 6 \tab 10 \tab 100 \tab 52 \tab 48 \tab 1 \tab 3 \tab 11 \tab 22 \tab 18 \tab 36 \tab 16 \tab 2 \tab 5 \tab 4 \tab 35 \tab 8 \tab 3 \tab 1 \tab 3 \tab 49 \tab 62 \tab 30 \tab 26 \tab 5 \tab 11 \tab 21 \tab 72 \tab 20 \tab 3 \tab 2 \tab 1 \tab 3 \tab 27 \tab 8 \tab 1 \tab 11 \tab 21 \tab 9 \tab 8 \tab 38 \tab 15 \tab 4 \tab 6 \tab - \tab 15 \tab 5 \tab 19 \tab 37 \tab 6 \tab 6 \tab 3 \tab 5 \tab 5 \tab - \tab 5 \tab 3 \tab 2 \tab 4 \tab 2 \tab 4 \tab 6 \tab - \tab 37 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab 4 \tab 35 \tab 3 \tab - \tab 1 \tab 7 \tab 9 \tab 1 \tab 7 \tab 3 \tab 24 \tab 13 \tab 1 \tab 2 \tab 16 \tab 9 \tab 1 \tab 1 \tab 58 \tab 14 \tab 4 \tab 2 \tab 11 \tab 2 \tab 4 \tab 6 \tab 43 \tab 4 \tab 49 \tab 8 \tab 4 \tab 4 \tab 1 \tab 2 \tab 11 \tab - \tab 13 \tab 1 \tab - \tab 8 \tab 27 \tab 118 \tab 5 \tab 6 \tab 1 \tab 4 \tab 1 \tab 6 \tab 10 \tab - \tab 8 \tab 2 \tab - \tab 11 \tab 3 \tab 1 \tab 6 \tab - \tab 5 \tab 9 \tab 6 \tab 2 \tab - \tab 3 \tab 7 \tab 2 \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 5 \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab 6 \tab 3 \tab 1 \tab 5 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 4 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 7 \tab 14 \tab 15 \tab 15 \tab 19 \tab 34 \tab 8 \tab 36 \tab 20 \tab 23 \tab 89 \tab 3 \tab 12 \tab 6 \tab 9 \tab 14 \tab 3 \tab 58 \tab 45 \tab 4 \tab 9 \tab 3 \tab 12 \tab 2 \tab 5 \tab 4 \tab 17 \tab 8 \tab 2 \tab 1 \tab 58 \tab 29 \tab 27 \tab 38 \tab 5 \tab 10 \tab 14 \tab 18 \tab 18 \tab 18 \tab 17 \tab 7 \tab - \tab 5 \tab 24 \tab 14 \tab 6 \tab 2 \tab 10 \tab 8 \tab 4 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 6 \tab - \tab - \tab 22 \tab 7 \tab 9 \tab 7 \tab 10 \tab 68 \tab 17 \tab 1 \tab 9 \tab 7 \tab 9 \tab 23 \tab 53 \tab 8 \tab 9 \tab 13 \tab 20 \tab 1 \tab - \tab 12 \tab 5 \tab 29 \tab 15 \tab 10 \tab 6 \tab 119 \tab 35 \tab 37 \tab 6 \tab 5 \tab 6 \tab 16 \tab 11 \tab 53 \tab 10 \tab 2 \tab 2 \tab 3 \tab 22 \tab 12 \tab 4 \tab 1 \tab 5 \tab 53 \tab 62 \tab 32 \tab 20 \tab 4 \tab 17 \tab 17 \tab 72 \tab 22 \tab 2 \tab 2 \tab 4 \tab 3 \tab 29 \tab 3 \tab 2 \tab 3 \tab 10 \tab 18 \tab 9 \tab 39 \tab 16 \tab 2 \tab 3 \tab 1 \tab 10 \tab 4 \tab 13 \tab 35 \tab 5 \tab 6 \tab 6 \tab 6 \tab 1 \tab - \tab 3 \tab 4 \tab 1 \tab 1 \tab 4 \tab 4 \tab 2 \tab 1 \tab 28 \tab 2 \tab 2 \tab 1 \tab 5 \tab - \tab 2 \tab 25 \tab 3 \tab - \tab - \tab 4 \tab 10 \tab - \tab 1 \tab 1 \tab 8 \tab 13 \tab 1 \tab 3 \tab 12 \tab 12 \tab 2 \tab - \tab 59 \tab 9 \tab 8 \tab 1 \tab 11 \tab - \tab 5 \tab 7 \tab 38 \tab - \tab 30 \tab 4 \tab 8 \tab 9 \tab 1 \tab - \tab 12 \tab - \tab 19 \tab 3 \tab 3 \tab 9 \tab 17 \tab 159 \tab 7 \tab 5 \tab - \tab 2 \tab - \tab 3 \tab 7 \tab 1 \tab 5 \tab - \tab 2 \tab 12 \tab 1 \tab - \tab 9 \tab - \tab 4 \tab 11 \tab 3 \tab 5 \tab - \tab - \tab 5 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab - \tab 4 \tab 2 \tab 5 \tab 1 \tab - \tab 2 \tab 6 \tab - \tab 2 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 8 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab 8 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 9 \tab 15 \tab 22 \tab 17 \tab 9 \tab 37 \tab 5 \tab 41 \tab 18 \tab 24 \tab 96 \tab 3 \tab 10 \tab 7 \tab 8 \tab 11 \tab 3 \tab 54 \tab 36 \tab 16 \tab 6 \tab 5 \tab 8 \tab 1 \tab 3 \tab 2 \tab 14 \tab 5 \tab 3 \tab 2 \tab 48 \tab 29 \tab 33 \tab 56 \tab 12 \tab 10 \tab 10 \tab 38 \tab 12 \tab 16 \tab 11 \tab 6 \tab - \tab 2 \tab 33 \tab 12 \tab 5 \tab 3 \tab 4 \tab 7 \tab 2 \tab 3 \tab 2 \tab 3 \tab 3 \tab 2 \tab 4 \tab - \tab 2 \tab 25 \tab 11 \tab 8 \tab 8 \tab 13 \tab 75 \tab 14 \tab 1 \tab 9 \tab 7 \tab 6 \tab 19 \tab 40 \tab 11 \tab 9 \tab 12 \tab 26 \tab 1 \tab 1 \tab 11 \tab 2 \tab 17 \tab 23 \tab 10 \tab 10 \tab 100 \tab 33 \tab 36 \tab 4 \tab 5 \tab 7 \tab 22 \tab 15 \tab 55 \tab 5 \tab 4 \tab 3 \tab 7 \tab 18 \tab 5 \tab 1 \tab 4 \tab 3 \tab 37 \tab 74 \tab 30 \tab 22 \tab 3 \tab 13 \tab 19 \tab 73 \tab 13 \tab 2 \tab 1 \tab 5 \tab 1 \tab 19 \tab 4 \tab 4 \tab 3 \tab 17 \tab 15 \tab 10 \tab 31 \tab 17 \tab 2 \tab 3 \tab 3 \tab 10 \tab 1 \tab 22 \tab 30 \tab 2 \tab 7 \tab 5 \tab 13 \tab 3 \tab - \tab 4 \tab 2 \tab 1 \tab 3 \tab 6 \tab 2 \tab 1 \tab 1 \tab 19 \tab 2 \tab - \tab 2 \tab 4 \tab 1 \tab 4 \tab 28 \tab - \tab - \tab - \tab 7 \tab 12 \tab 1 \tab 1 \tab 2 \tab 18 \tab 11 \tab 4 \tab 5 \tab 11 \tab 8 \tab 1 \tab - \tab 50 \tab 9 \tab 11 \tab 2 \tab 6 \tab - \tab 4 \tab 5 \tab 50 \tab 3 \tab 49 \tab 8 \tab 4 \tab 6 \tab 1 \tab - \tab 13 \tab - \tab 23 \tab 1 \tab 2 \tab 10 \tab 39 \tab 143 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab 6 \tab - \tab 3 \tab 11 \tab 1 \tab - \tab 8 \tab - \tab 3 \tab 14 \tab 5 \tab 4 \tab 1 \tab 2 \tab 3 \tab 2 \tab 4 \tab - \tab 1 \tab - \tab - \tab 4 \tab 4 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 5 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 2817 \cr
   2002 \tab 12 \tab 13 \tab 13 \tab 6 \tab 18 \tab 27 \tab 3 \tab 35 \tab 23 \tab 18 \tab 87 \tab 1 \tab 8 \tab 3 \tab 6 \tab 15 \tab 5 \tab 58 \tab 39 \tab 7 \tab 5 \tab 6 \tab 7 \tab - \tab 9 \tab 1 \tab 14 \tab 9 \tab 1 \tab - \tab 35 \tab 25 \tab 35 \tab 50 \tab 9 \tab 4 \tab 7 \tab 26 \tab 23 \tab 20 \tab 11 \tab 5 \tab - \tab 2 \tab 42 \tab 12 \tab 9 \tab 7 \tab 4 \tab 7 \tab 4 \tab 2 \tab 5 \tab 4 \tab 8 \tab 5 \tab 7 \tab - \tab 1 \tab 20 \tab 9 \tab 14 \tab 5 \tab 24 \tab 65 \tab 16 \tab 2 \tab 7 \tab 5 \tab 12 \tab 19 \tab 69 \tab 11 \tab 6 \tab 11 \tab 18 \tab 1 \tab 5 \tab 11 \tab 1 \tab 34 \tab 20 \tab 12 \tab 7 \tab 100 \tab 32 \tab 34 \tab 2 \tab 6 \tab 4 \tab 22 \tab 14 \tab 48 \tab 6 \tab 4 \tab 3 \tab 4 \tab 24 \tab 7 \tab - \tab - \tab 6 \tab 51 \tab 62 \tab 40 \tab 18 \tab 11 \tab 12 \tab 22 \tab 81 \tab 14 \tab 4 \tab 7 \tab 3 \tab - \tab 25 \tab 4 \tab 3 \tab - \tab 11 \tab 11 \tab 10 \tab 28 \tab 9 \tab 5 \tab 3 \tab - \tab 12 \tab 2 \tab 29 \tab 27 \tab 2 \tab 10 \tab 4 \tab 14 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 5 \tab 3 \tab 2 \tab - \tab - \tab 24 \tab 1 \tab 6 \tab - \tab 2 \tab 1 \tab 2 \tab 22 \tab 2 \tab 1 \tab - \tab 8 \tab 12 \tab 2 \tab 2 \tab 2 \tab 12 \tab 8 \tab - \tab 1 \tab 7 \tab 10 \tab - \tab 1 \tab 57 \tab 12 \tab 9 \tab 1 \tab 8 \tab 2 \tab 5 \tab 5 \tab 33 \tab 2 \tab 47 \tab 4 \tab 8 \tab 2 \tab 3 \tab 1 \tab 16 \tab 1 \tab 20 \tab - \tab 4 \tab 8 \tab 28 \tab 112 \tab - \tab 7 \tab 1 \tab 3 \tab - \tab 4 \tab 4 \tab - \tab 12 \tab - \tab 1 \tab 14 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 13 \tab 5 \tab 5 \tab - \tab 1 \tab 5 \tab - \tab 3 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 4 \tab 8 \tab - \tab 2 \tab 5 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 7 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 5 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2765 \cr
   2004 \tab 11 \tab 17 \tab 22 \tab 19 \tab 18 \tab 28 \tab 6 \tab 30 \tab 26 \tab 17 \tab 83 \tab 3 \tab 5 \tab 5 \tab 10 \tab 23 \tab 6 \tab 64 \tab 42 \tab 8 \tab 10 \tab 6 \tab 6 \tab 2 \tab 9 \tab - \tab 18 \tab 13 \tab 2 \tab 1 \tab 31 \tab 22 \tab 37 \tab 59 \tab 14 \tab 8 \tab 14 \tab 36 \tab 20 \tab 15 \tab 15 \tab 9 \tab - \tab 3 \tab 41 \tab 7 \tab 5 \tab 5 \tab 8 \tab 4 \tab 3 \tab 2 \tab 4 \tab 4 \tab 9 \tab 2 \tab 4 \tab - \tab 1 \tab 18 \tab 17 \tab 12 \tab 4 \tab 21 \tab 91 \tab 24 \tab 2 \tab 13 \tab 6 \tab 13 \tab 13 \tab 59 \tab 8 \tab 6 \tab 7 \tab 23 \tab - \tab 2 \tab 12 \tab 3 \tab 17 \tab 14 \tab 9 \tab 16 \tab 100 \tab 41 \tab 21 \tab 9 \tab 8 \tab 8 \tab 12 \tab 17 \tab 38 \tab 5 \tab 1 \tab 2 \tab 3 \tab 15 \tab 3 \tab 1 \tab - \tab 6 \tab 39 \tab 69 \tab 20 \tab 18 \tab 8 \tab 8 \tab 16 \tab 73 \tab 14 \tab 4 \tab 5 \tab 4 \tab - \tab 32 \tab 7 \tab 1 \tab 3 \tab 11 \tab 9 \tab 6 \tab 20 \tab 13 \tab 3 \tab - \tab 1 \tab 10 \tab 8 \tab 20 \tab 37 \tab 4 \tab 10 \tab - \tab 9 \tab 4 \tab - \tab 5 \tab 5 \tab 1 \tab 3 \tab 5 \tab 2 \tab 2 \tab 1 \tab 14 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab 30 \tab 2 \tab - \tab 1 \tab 8 \tab 8 \tab - \tab 2 \tab 1 \tab 15 \tab 4 \tab - \tab 2 \tab 3 \tab 6 \tab 3 \tab - \tab 58 \tab 13 \tab 13 \tab 1 \tab 10 \tab - \tab 1 \tab 3 \tab 31 \tab 7 \tab 39 \tab 1 \tab 7 \tab 2 \tab 2 \tab 2 \tab 15 \tab - \tab 25 \tab - \tab 2 \tab 11 \tab 26 \tab 154 \tab - \tab 4 \tab - \tab 6 \tab 2 \tab 9 \tab 6 \tab - \tab 14 \tab 1 \tab - \tab 18 \tab - \tab 1 \tab 6 \tab - \tab 3 \tab 7 \tab 6 \tab 13 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab 2 \tab 4 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 4 \tab 5 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 2812 \cr
   2006 \tab 29 \tab 18 \tab 39 \tab 27 \tab 27 \tab 41 \tab 18 \tab 52 \tab 33 \tab 27 \tab 121 \tab 1 \tab 7 \tab 4 \tab 12 \tab 20 \tab 1 \tab 88 \tab 76 \tab 13 \tab 16 \tab 10 \tab 11 \tab 10 \tab 13 \tab 5 \tab 25 \tab 18 \tab 2 \tab - \tab 51 \tab 47 \tab 42 \tab 114 \tab 18 \tab 13 \tab 16 \tab 41 \tab 22 \tab 25 \tab 26 \tab 10 \tab 1 \tab 4 \tab 56 \tab 19 \tab 13 \tab 8 \tab 14 \tab 11 \tab 2 \tab 12 \tab 5 \tab 5 \tab 2 \tab 1 \tab 7 \tab - \tab 3 \tab 30 \tab 18 \tab 15 \tab 13 \tab 36 \tab 103 \tab 30 \tab 1 \tab 20 \tab 10 \tab 21 \tab 34 \tab 96 \tab 16 \tab 9 \tab 20 \tab 26 \tab 2 \tab 4 \tab 31 \tab 4 \tab 33 \tab 35 \tab 16 \tab 10 \tab 141 \tab 64 \tab 57 \tab 13 \tab 11 \tab 10 \tab 24 \tab 32 \tab 71 \tab 26 \tab 2 \tab 6 \tab 4 \tab 27 \tab 8 \tab 1 \tab 2 \tab 9 \tab 81 \tab 92 \tab 48 \tab 42 \tab 9 \tab 25 \tab 27 \tab 94 \tab 30 \tab 8 \tab 4 \tab 4 \tab 2 \tab 41 \tab 7 \tab 9 \tab 6 \tab 20 \tab 20 \tab 13 \tab 40 \tab 18 \tab 6 \tab 4 \tab 2 \tab 20 \tab 3 \tab 28 \tab 59 \tab 3 \tab 11 \tab 3 \tab 21 \tab 4 \tab - \tab 8 \tab 5 \tab 3 \tab 10 \tab 6 \tab 4 \tab 7 \tab 2 \tab 33 \tab 1 \tab 7 \tab 2 \tab 3 \tab 1 \tab 4 \tab 40 \tab 1 \tab - \tab - \tab 10 \tab 10 \tab 1 \tab 5 \tab - \tab 20 \tab 7 \tab - \tab 3 \tab 10 \tab 16 \tab 1 \tab 1 \tab 99 \tab 11 \tab 18 \tab 4 \tab 13 \tab - \tab 6 \tab 4 \tab 74 \tab 13 \tab 71 \tab 8 \tab 13 \tab 5 \tab 1 \tab 3 \tab 25 \tab 1 \tab 57 \tab 2 \tab 4 \tab 14 \tab 42 \tab 223 \tab - \tab 6 \tab 3 \tab 3 \tab 2 \tab 9 \tab 1 \tab 1 \tab 15 \tab - \tab 4 \tab 25 \tab 2 \tab 3 \tab 8 \tab 1 \tab 2 \tab 18 \tab 12 \tab 16 \tab 1 \tab 4 \tab 9 \tab 7 \tab 3 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 6 \tab 5 \tab 1 \tab 1 \tab 6 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab 5 \tab 1 \tab 10 \tab - \tab 1 \tab 7 \tab 4 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 5 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 4 \tab 3 \tab 5 \tab 4 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 11 \tab 14 \tab 6 \tab 11 \tab 15 \tab 19 \tab 5 \tab 22 \tab 13 \tab 21 \tab 65 \tab 3 \tab 3 \tab 8 \tab 3 \tab 18 \tab 3 \tab 38 \tab 41 \tab 7 \tab 3 \tab 7 \tab 5 \tab 7 \tab 4 \tab 1 \tab 13 \tab 3 \tab 2 \tab 3 \tab 33 \tab 23 \tab 24 \tab 31 \tab 10 \tab 4 \tab 8 \tab 23 \tab 17 \tab 12 \tab 10 \tab 3 \tab - \tab 4 \tab 31 \tab 6 \tab 4 \tab 3 \tab 8 \tab 3 \tab - \tab 3 \tab 6 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab 17 \tab 6 \tab 14 \tab 4 \tab 8 \tab 55 \tab 6 \tab 1 \tab 7 \tab 3 \tab 7 \tab 10 \tab 48 \tab 10 \tab 4 \tab 4 \tab 21 \tab 2 \tab 1 \tab 10 \tab 1 \tab 11 \tab 14 \tab 8 \tab 8 \tab 61 \tab 15 \tab 27 \tab 3 \tab 6 \tab 6 \tab 10 \tab 12 \tab 30 \tab 3 \tab - \tab 2 \tab 3 \tab 16 \tab 5 \tab 3 \tab - \tab 7 \tab 38 \tab 53 \tab 21 \tab 18 \tab 4 \tab 8 \tab 13 \tab 35 \tab 13 \tab - \tab 4 \tab 2 \tab 3 \tab 26 \tab 1 \tab 4 \tab 3 \tab 4 \tab 6 \tab 4 \tab 13 \tab 13 \tab 1 \tab 4 \tab 2 \tab 6 \tab 1 \tab 20 \tab 23 \tab 1 \tab 2 \tab 1 \tab 5 \tab 2 \tab - \tab 8 \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 16 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 19 \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab 4 \tab 3 \tab 1 \tab 14 \tab 5 \tab - \tab 2 \tab 3 \tab 4 \tab - \tab - \tab 53 \tab 6 \tab 4 \tab 2 \tab 12 \tab - \tab - \tab 2 \tab 22 \tab 4 \tab 25 \tab 3 \tab 5 \tab 3 \tab 2 \tab 4 \tab 17 \tab - \tab 24 \tab 3 \tab 4 \tab 2 \tab 21 \tab 72 \tab - \tab 4 \tab 2 \tab - \tab 1 \tab 2 \tab 5 \tab - \tab 5 \tab - \tab - \tab 16 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 8 \tab 9 \tab 7 \tab 1 \tab 3 \tab 4 \tab 2 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 4 \tab 1 \tab 4 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab 6 \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 3 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 23 \tab 13 \tab 26 \tab 10 \tab 9 \tab 16 \tab 4 \tab 28 \tab 23 \tab 12 \tab 58 \tab 1 \tab 2 \tab 8 \tab 8 \tab 14 \tab 1 \tab 44 \tab 22 \tab 3 \tab 2 \tab 1 \tab 4 \tab 1 \tab 9 \tab - \tab 12 \tab 2 \tab 3 \tab 1 \tab 32 \tab 24 \tab 16 \tab 47 \tab 11 \tab 9 \tab 12 \tab 18 \tab 13 \tab 8 \tab 10 \tab 10 \tab - \tab 1 \tab 20 \tab 7 \tab 5 \tab 4 \tab 8 \tab - \tab 2 \tab 2 \tab 2 \tab 3 \tab 3 \tab 1 \tab 5 \tab - \tab 2 \tab 8 \tab 9 \tab 7 \tab 6 \tab 13 \tab 47 \tab 9 \tab 4 \tab 11 \tab 1 \tab 10 \tab 14 \tab 49 \tab 4 \tab 6 \tab 9 \tab 14 \tab 1 \tab 3 \tab 12 \tab - \tab 18 \tab 19 \tab 3 \tab 4 \tab 53 \tab 17 \tab 18 \tab 5 \tab 1 \tab 2 \tab 8 \tab 12 \tab 37 \tab 4 \tab - \tab 3 \tab 1 \tab 9 \tab 4 \tab 1 \tab - \tab 8 \tab 40 \tab 60 \tab 27 \tab 17 \tab 2 \tab 15 \tab 11 \tab 54 \tab 15 \tab 2 \tab 4 \tab 3 \tab 1 \tab 21 \tab 2 \tab 4 \tab 2 \tab 9 \tab 8 \tab 8 \tab 16 \tab 16 \tab 3 \tab 2 \tab - \tab 8 \tab - \tab 9 \tab 19 \tab 2 \tab 8 \tab 2 \tab 5 \tab 5 \tab - \tab - \tab 2 \tab - \tab - \tab 4 \tab 2 \tab - \tab - \tab 12 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab 11 \tab - \tab - \tab - \tab 2 \tab 5 \tab 1 \tab 2 \tab 1 \tab 9 \tab 5 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab 41 \tab 6 \tab 8 \tab 2 \tab 8 \tab 2 \tab 1 \tab 3 \tab 25 \tab - \tab 38 \tab 1 \tab 6 \tab 6 \tab - \tab 1 \tab 12 \tab - \tab 17 \tab 1 \tab 1 \tab 9 \tab 26 \tab 117 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 7 \tab 5 \tab - \tab 7 \tab - \tab 2 \tab 20 \tab 1 \tab 2 \tab 5 \tab - \tab 1 \tab 16 \tab 2 \tab 6 \tab - \tab 4 \tab 5 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 30 \tab 9 \tab 21 \tab 5 \tab 5 \tab 20 \tab 2 \tab 23 \tab 24 \tab 14 \tab 47 \tab 1 \tab 7 \tab 5 \tab 5 \tab 7 \tab 3 \tab 49 \tab 28 \tab 9 \tab 6 \tab 3 \tab 6 \tab 2 \tab 6 \tab - \tab 7 \tab 7 \tab - \tab - \tab 46 \tab 24 \tab 29 \tab 41 \tab 12 \tab 9 \tab 7 \tab 27 \tab 11 \tab 6 \tab 7 \tab 2 \tab 1 \tab 1 \tab 23 \tab 9 \tab 2 \tab 2 \tab 4 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 3 \tab 4 \tab 7 \tab - \tab - \tab 7 \tab 14 \tab 15 \tab 5 \tab 12 \tab 46 \tab 19 \tab - \tab 5 \tab 2 \tab 12 \tab 10 \tab 46 \tab 6 \tab 9 \tab 7 \tab 17 \tab - \tab 2 \tab 7 \tab - \tab 15 \tab 11 \tab 2 \tab 7 \tab 51 \tab 17 \tab 21 \tab 5 \tab 1 \tab 1 \tab 6 \tab 11 \tab 32 \tab 6 \tab - \tab 2 \tab 1 \tab 15 \tab - \tab - \tab 2 \tab 3 \tab 43 \tab 45 \tab 28 \tab 19 \tab 5 \tab 7 \tab 9 \tab 42 \tab 29 \tab 2 \tab 2 \tab 3 \tab 2 \tab 9 \tab 4 \tab 1 \tab 5 \tab 9 \tab 7 \tab 4 \tab 14 \tab 4 \tab 1 \tab 6 \tab - \tab 7 \tab 1 \tab 8 \tab 14 \tab 3 \tab 5 \tab 1 \tab 8 \tab 1 \tab - \tab 3 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 10 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 3 \tab 26 \tab - \tab 1 \tab - \tab 7 \tab 2 \tab 1 \tab - \tab - \tab 6 \tab 6 \tab - \tab - \tab 5 \tab 8 \tab - \tab - \tab 44 \tab 6 \tab 6 \tab 2 \tab 4 \tab - \tab 1 \tab 1 \tab 31 \tab 3 \tab 42 \tab 4 \tab 5 \tab 3 \tab 1 \tab - \tab 7 \tab - \tab 27 \tab - \tab - \tab 6 \tab 21 \tab 103 \tab - \tab 6 \tab - \tab 1 \tab 1 \tab 9 \tab 2 \tab - \tab 6 \tab - \tab 2 \tab 10 \tab - \tab - \tab 4 \tab - \tab 3 \tab 15 \tab 4 \tab 5 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 5 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 33 \tab 19 \tab 24 \tab 7 \tab 6 \tab 26 \tab 4 \tab 40 \tab 14 \tab 28 \tab 79 \tab 3 \tab 14 \tab 7 \tab 12 \tab 14 \tab 1 \tab 32 \tab 36 \tab 8 \tab 2 \tab 7 \tab 7 \tab 2 \tab 9 \tab - \tab 17 \tab 13 \tab 2 \tab 2 \tab 32 \tab 22 \tab 29 \tab 51 \tab 13 \tab 7 \tab 15 \tab 31 \tab 24 \tab 11 \tab 15 \tab 12 \tab 1 \tab 7 \tab 32 \tab 11 \tab 3 \tab 3 \tab 5 \tab 4 \tab 2 \tab 4 \tab 2 \tab 1 \tab 4 \tab 4 \tab 7 \tab - \tab - \tab 15 \tab 10 \tab 9 \tab 6 \tab 24 \tab 82 \tab 10 \tab 2 \tab 9 \tab 3 \tab 10 \tab 22 \tab 44 \tab 7 \tab 7 \tab 13 \tab 22 \tab 1 \tab 2 \tab 8 \tab - \tab 30 \tab 12 \tab 3 \tab 8 \tab 59 \tab 23 \tab 35 \tab 4 \tab 4 \tab 2 \tab 9 \tab 21 \tab 40 \tab 5 \tab 3 \tab 3 \tab 4 \tab 16 \tab 5 \tab 4 \tab 1 \tab 8 \tab 42 \tab 78 \tab 19 \tab 23 \tab 5 \tab 15 \tab 9 \tab 70 \tab 20 \tab 2 \tab 1 \tab 7 \tab 1 \tab 16 \tab 2 \tab 5 \tab 2 \tab 10 \tab 9 \tab 9 \tab 23 \tab 5 \tab 2 \tab 2 \tab 1 \tab 10 \tab 2 \tab 17 \tab 32 \tab 5 \tab 5 \tab 3 \tab 9 \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 7 \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 4 \tab 30 \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 7 \tab 5 \tab 1 \tab - \tab 8 \tab 7 \tab 1 \tab - \tab 57 \tab 11 \tab 7 \tab 2 \tab 7 \tab - \tab 1 \tab 4 \tab 42 \tab 2 \tab 42 \tab - \tab 11 \tab 6 \tab 1 \tab 6 \tab 12 \tab - \tab 29 \tab 3 \tab - \tab 8 \tab 28 \tab 111 \tab - \tab 7 \tab 1 \tab - \tab - \tab 9 \tab 4 \tab - \tab 10 \tab - \tab 1 \tab 20 \tab 3 \tab 1 \tab 4 \tab - \tab 2 \tab 8 \tab 12 \tab 1 \tab - \tab 2 \tab 5 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 6 \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab - \tab 7 \tab 3 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 5 \tab - \tab 3 \tab 1 \tab 4 \tab - \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 2538 \cr
   2016 \tab 18 \tab 23 \tab 23 \tab 3 \tab 11 \tab 23 \tab 6 \tab 31 \tab 36 \tab 17 \tab 73 \tab 5 \tab 8 \tab 5 \tab 20 \tab 18 \tab 5 \tab 64 \tab 45 \tab 14 \tab 9 \tab 8 \tab 5 \tab 2 \tab 7 \tab 2 \tab 10 \tab 11 \tab 1 \tab - \tab 55 \tab 35 \tab 34 \tab 54 \tab 17 \tab 6 \tab 13 \tab 29 \tab 24 \tab 21 \tab 14 \tab 14 \tab 1 \tab 6 \tab 37 \tab 10 \tab 4 \tab 7 \tab 4 \tab 3 \tab 2 \tab 7 \tab 1 \tab 5 \tab 4 \tab 4 \tab 8 \tab - \tab 2 \tab 17 \tab 20 \tab 12 \tab 11 \tab 21 \tab 81 \tab 23 \tab 2 \tab 10 \tab 3 \tab 14 \tab 5 \tab 73 \tab 6 \tab 9 \tab 5 \tab 30 \tab 1 \tab 2 \tab 17 \tab 2 \tab 29 \tab 10 \tab 6 \tab 8 \tab 69 \tab 32 \tab 17 \tab 5 \tab 5 \tab 5 \tab 14 \tab 24 \tab 44 \tab 8 \tab 1 \tab 3 \tab - \tab 23 \tab 5 \tab 4 \tab 2 \tab 5 \tab 63 \tab 61 \tab 32 \tab 31 \tab 6 \tab 13 \tab 16 \tab 61 \tab 23 \tab 3 \tab 2 \tab 2 \tab 1 \tab 20 \tab 5 \tab 4 \tab 4 \tab 6 \tab 15 \tab 10 \tab 23 \tab 16 \tab 3 \tab 4 \tab - \tab 10 \tab 1 \tab 14 \tab 25 \tab 4 \tab 3 \tab 5 \tab 6 \tab 1 \tab - \tab 6 \tab 7 \tab 2 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab 14 \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 30 \tab - \tab - \tab 1 \tab 3 \tab 6 \tab - \tab - \tab - \tab 6 \tab 8 \tab 1 \tab 1 \tab 7 \tab 10 \tab 1 \tab - \tab 58 \tab 14 \tab 11 \tab 1 \tab 10 \tab - \tab 5 \tab - \tab 34 \tab 7 \tab 49 \tab 4 \tab 8 \tab 8 \tab 1 \tab 4 \tab 17 \tab 1 \tab 37 \tab - \tab - \tab 5 \tab 31 \tab 119 \tab - \tab 7 \tab 3 \tab 2 \tab - \tab 6 \tab 2 \tab 1 \tab 11 \tab 1 \tab 3 \tab 16 \tab - \tab 3 \tab 4 \tab - \tab 4 \tab 32 \tab 15 \tab 15 \tab - \tab 3 \tab 5 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab 4 \tab 3 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 3 \tab 2 \tab 3 \tab 4 \tab 1 \tab 7 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 5 \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab 5 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab 3 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2867 \cr
   2018 \tab 13 \tab 21 \tab 21 \tab 9 \tab 18 \tab 23 \tab 3 \tab 26 \tab 15 \tab 26 \tab 63 \tab 1 \tab 5 \tab 4 \tab 10 \tab 15 \tab 1 \tab 31 \tab 31 \tab 9 \tab 12 \tab 5 \tab 4 \tab 4 \tab 6 \tab 1 \tab 14 \tab 5 \tab 4 \tab 1 \tab 57 \tab 30 \tab 26 \tab 40 \tab 8 \tab 7 \tab 17 \tab 26 \tab 14 \tab 17 \tab 16 \tab 7 \tab - \tab 3 \tab 30 \tab 11 \tab 3 \tab 3 \tab 6 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 6 \tab 4 \tab - \tab 3 \tab 17 \tab 16 \tab 11 \tab 10 \tab 24 \tab 69 \tab 12 \tab 1 \tab 15 \tab 4 \tab 13 \tab 17 \tab 29 \tab 12 \tab 8 \tab 13 \tab 31 \tab - \tab - \tab 10 \tab 1 \tab 18 \tab 9 \tab 5 \tab 7 \tab 51 \tab 14 \tab 38 \tab 4 \tab 6 \tab 2 \tab 17 \tab 22 \tab 24 \tab 10 \tab 2 \tab 1 \tab 1 \tab 24 \tab 2 \tab 1 \tab 3 \tab 1 \tab 55 \tab 52 \tab 21 \tab 17 \tab 11 \tab 12 \tab 8 \tab 52 \tab 24 \tab 1 \tab 2 \tab 1 \tab - \tab 12 \tab 7 \tab 1 \tab 4 \tab 7 \tab 12 \tab 2 \tab 16 \tab 9 \tab 5 \tab 2 \tab - \tab 5 \tab 3 \tab 14 \tab 27 \tab 7 \tab 11 \tab 1 \tab 8 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 5 \tab 2 \tab 6 \tab 3 \tab - \tab 2 \tab 4 \tab 14 \tab - \tab - \tab - \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 8 \tab 2 \tab - \tab 2 \tab 3 \tab 12 \tab 1 \tab - \tab 50 \tab 13 \tab 4 \tab 1 \tab 8 \tab 1 \tab 1 \tab 3 \tab 33 \tab 3 \tab 37 \tab 7 \tab 6 \tab 2 \tab - \tab 1 \tab 16 \tab - \tab 27 \tab 2 \tab 1 \tab 12 \tab 35 \tab 100 \tab - \tab 7 \tab 2 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 4 \tab - \tab 3 \tab 25 \tab - \tab 2 \tab 2 \tab - \tab 4 \tab 13 \tab 5 \tab 3 \tab - \tab 1 \tab 5 \tab - \tab 4 \tab 3 \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 6 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 7 \tab 3 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 460 \tab 335 \tab 421 \tab 575 \tab 339 \tab 603 \tab 132 \tab 753 \tab 457 \tab 385 \tab 1794 \tab 63 \tab 152 \tab 92 \tab 158 \tab 280 \tab 57 \tab 1019 \tab 614 \tab 156 \tab 162 \tab 136 \tab 160 \tab 51 \tab 149 \tab 38 \tab 383 \tab 187 \tab 53 \tab 21 \tab 960 \tab 565 \tab 792 \tab 1239 \tab 221 \tab 143 \tab 240 \tab 602 \tab 290 \tab 260 \tab 285 \tab 162 \tab 14 \tab 66 \tab 640 \tab 210 \tab 134 \tab 85 \tab 151 \tab 89 \tab 66 \tab 82 \tab 95 \tab 118 \tab 80 \tab 53 \tab 113 \tab 3 \tab 25 \tab 536 \tab 281 \tab 226 \tab 154 \tab 315 \tab 1566 \tab 336 \tab 29 \tab 165 \tab 96 \tab 289 \tab 346 \tab 1019 \tab 200 \tab 154 \tab 203 \tab 423 \tab 22 \tab 30 \tab 267 \tab 51 \tab 618 \tab 525 \tab 223 \tab 187 \tab 2563 \tab 1043 \tab 804 \tab 137 \tab 103 \tab 185 \tab 391 \tab 337 \tab 957 \tab 207 \tab 48 \tab 54 \tab 60 \tab 422 \tab 252 \tab 37 \tab 35 \tab 123 \tab 1057 \tab 1526 \tab 585 \tab 581 \tab 128 \tab 288 \tab 338 \tab 1615 \tab 362 \tab 54 \tab 73 \tab 101 \tab 40 \tab 605 \tab 101 \tab 69 \tab 68 \tab 267 \tab 304 \tab 190 \tab 604 \tab 342 \tab 94 \tab 86 \tab 19 \tab 318 \tab 87 \tab 471 \tab 702 \tab 85 \tab 130 \tab 66 \tab 225 \tab 94 \tab 7 \tab 138 \tab 90 \tab 44 \tab 72 \tab 128 \tab 64 \tab 39 \tab 16 \tab 549 \tab 39 \tab 77 \tab 61 \tab 51 \tab 14 \tab 78 \tab 629 \tab 41 \tab 8 \tab 11 \tab 186 \tab 209 \tab 33 \tab 81 \tab 52 \tab 561 \tab 234 \tab 47 \tab 71 \tab 219 \tab 216 \tab 36 \tab 17 \tab 1310 \tab 218 \tab 227 \tab 76 \tab 184 \tab 15 \tab 100 \tab 62 \tab 1139 \tab 96 \tab 962 \tab 82 \tab 139 \tab 94 \tab 32 \tab 44 \tab 399 \tab 11 \tab 556 \tab 45 \tab 49 \tab 251 \tab 675 \tab 11478 \tab 863 \tab 98 \tab 27 \tab 56 \tab 23 \tab 108 \tab 103 \tab 14 \tab 119 \tab 10 \tab 39 \tab 268 \tab 25 \tab 25 \tab 121 \tab 17 \tab 64 \tab 206 \tab 137 \tab 122 \tab 8 \tab 59 \tab 112 \tab 32 \tab 57 \tab 32 \tab 69 \tab 30 \tab 17 \tab 27 \tab 42 \tab 63 \tab 89 \tab 15 \tab 36 \tab 44 \tab 18 \tab 26 \tab 52 \tab 31 \tab 12 \tab 23 \tab 17 \tab 13 \tab 116 \tab 27 \tab 9 \tab 59 \tab 25 \tab 18 \tab 20 \tab 19 \tab 64 \tab 13 \tab 17 \tab 28 \tab 6 \tab 10 \tab 2 \tab 38 \tab 76 \tab 21 \tab 22 \tab 11 \tab 13 \tab 25 \tab 34 \tab 52 \tab 43 \tab 11 \tab 26 \tab 11 \tab 19 \tab 18 \tab 32 \tab 10 \tab 10 \tab 16 \tab 41 \tab 2 \tab 2 \tab 5 \tab 1 \tab 1 \tab 10 \tab 8 \tab 5 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{9997} uncodeable
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
