% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_600.R
\name{israel}
\alias{israel}
\title{israel}
\source{
General Social Survey https://gss.norc.org
}
\description{
Liking for israel
}
\details{
Question 103. You will notice that the boxes on this card go from the highest position of "plus 5" for a country which you like very much, to the lowest position of "minus 5" for a country you dislike very much. How far up the scale or how far down the scale would you rate the following countries? g. Israel
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab not available in this year \tab +1 \tab +2 \tab +3 \tab +4 \tab +5 \tab -1 \tab -2 \tab -3 \tab -4 \tab -5 \tab dk, na, iap \tab iap \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 258 \tab 175 \tab 231 \tab 141 \tab 198 \tab 128 \tab 57 \tab 67 \tab 33 \tab 89 \tab 107 \tab - \tab 1484 \cr
   1975 \tab - \tab 275 \tab 169 \tab 205 \tab 123 \tab 146 \tab 164 \tab 82 \tab 77 \tab 48 \tab 76 \tab 125 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 270 \tab 173 \tab 204 \tab 151 \tab 180 \tab 149 \tab 71 \tab 64 \tab 41 \tab 69 \tab 158 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 290 \tab 211 \tab 250 \tab 147 \tab 207 \tab 215 \tab 121 \tab 92 \tab 62 \tab 95 \tab 170 \tab - \tab 1860 \cr
   1983 \tab - \tab 285 \tab 172 \tab 169 \tab 123 \tab 148 \tab 209 \tab 106 \tab 111 \tab 67 \tab 103 \tab 106 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 254 \tab 201 \tab 211 \tab 125 \tab 168 \tab 160 \tab 94 \tab 82 \tab 48 \tab 75 \tab 116 \tab - \tab 1534 \cr
   1986 \tab - \tab 285 \tab 180 \tab 175 \tab 128 \tab 165 \tab 144 \tab 84 \tab 73 \tab 53 \tab 104 \tab 79 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 155 \tab 91 \tab 101 \tab 60 \tab 96 \tab 130 \tab 77 \tab 77 \tab 30 \tab 94 \tab 570 \tab - \tab 1481 \cr
   1989 \tab - \tab 177 \tab 95 \tab 107 \tab 56 \tab 87 \tab 104 \tab 70 \tab 59 \tab 43 \tab 118 \tab 621 \tab - \tab 1537 \cr
   1990 \tab - \tab 182 \tab 93 \tab 86 \tab 48 \tab 75 \tab 128 \tab 69 \tab 45 \tab 30 \tab 93 \tab 523 \tab - \tab 1372 \cr
   1991 \tab - \tab 177 \tab 138 \tab 162 \tab 104 \tab 152 \tab 76 \tab 38 \tab 35 \tab 14 \tab 55 \tab 566 \tab - \tab 1517 \cr
   1993 \tab - \tab 227 \tab 104 \tab 133 \tab 73 \tab 105 \tab 123 \tab 71 \tab 59 \tab 41 \tab 66 \tab 604 \tab - \tab 1606 \cr
   1994 \tab - \tab 110 \tab 59 \tab 52 \tab 32 \tab 46 \tab 55 \tab 31 \tab 34 \tab 14 \tab 34 \tab 2525 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 47374 \tab 2945 \tab 1861 \tab 2086 \tab 1311 \tab 1773 \tab 1785 \tab 971 \tab 875 \tab 524 \tab 1071 \tab 6270 \tab 3544 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} +5
\item \code{1} +4
\item \code{2} +3
\item \code{3} +2
\item \code{4} +1
\item \code{5} -1
\item \code{6} -2
\item \code{7} -3
\item \code{8} -4
\item \code{9} -5
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
