% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{issp}
\alias{issp}
\title{issp}
\source{
General Social Survey https://gss.norc.org
}
\description{
Filter for issp cases
}
\details{
Question 1661. Status on International Social Survey Program Self-Completion Supplement.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab r did issp \tab r did not do issp \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab 1473 \cr
   1985 \tab 780 \tab 677 \tab 77 \tab 1534 \cr
   1986 \tab - \tab 1428 \tab 42 \tab 1470 \cr
   1987 \tab - \tab 1564 \tab 255 \tab 1819 \cr
   1988 \tab - \tab 1414 \tab 67 \tab 1481 \cr
   1989 \tab - \tab 1453 \tab 84 \tab 1537 \cr
   1990 \tab - \tab 1217 \tab 155 \tab 1372 \cr
   1991 \tab - \tab 1359 \tab 158 \tab 1517 \cr
   1993 \tab - \tab 1557 \tab 49 \tab 1606 \cr
   1994 \tab - \tab 2833 \tab 159 \tab 2992 \cr
   1996 \tab - \tab 2699 \tab 205 \tab 2904 \cr
   1998 \tab - \tab 2507 \tab 325 \tab 2832 \cr
   2000 \tab - \tab 2548 \tab 269 \tab 2817 \cr
   2002 \tab - \tab 2320 \tab 445 \tab 2765 \cr
   2004 \tab 1345 \tab 1216 \tab 251 \tab 2812 \cr
   2006 \tab 2992 \tab 1515 \tab 3 \tab 4510 \cr
   2008 \tab 658 \tab 1360 \tab 5 \tab 2023 \cr
   2010 \tab 614 \tab 1413 \tab 17 \tab 2044 \cr
   2012 \tab 672 \tab 1282 \tab 20 \tab 1974 \cr
   2014 \tab - \tab 2522 \tab 16 \tab 2538 \cr
   2016 \tab - \tab 2846 \tab 21 \tab 2867 \cr
   2018 \tab - \tab 2337 \tab 11 \tab 2348 \cr
   2021 \tab - \tab 3700 \tab 332 \tab 4032 \cr
   2022 \tab - \tab 2333 \tab 1211 \tab 3544 \cr
   Total \tab 24113 \tab 44100 \tab 4177 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} r did issp
\item \code{2} r did not do issp
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
