% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_61.R
\name{jobhour}
\alias{jobhour}
\title{jobhour}
\source{
General Social Survey https://gss.norc.org
}
\description{
Short working hours
}
\details{
Question 183a. Would you please look at this card and tell me which one thing on this list you would most prefer in a job? b. Which comes next? c. Which is third most important? d. Which is fourth most important? WORKING HOURS ARE SHORT, LOTS OF FREE TIME
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab fifth \tab fourth \tab most impt \tab no answer \tab second \tab third \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 684 \tab 370 \tab 70 \tab 42 \tab 147 \tab 191 \tab 1504 \cr
   1974 \tab - \tab 621 \tab 393 \tab 71 \tab 30 \tab 166 \tab 203 \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 759 \tab 325 \tab 58 \tab 50 \tab 133 \tab 174 \tab 1499 \cr
   1977 \tab - \tab 768 \tab 368 \tab 66 \tab 50 \tab 116 \tab 162 \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 740 \tab 371 \tab 47 \tab 32 \tab 122 \tab 156 \tab 1468 \cr
   1982 \tab 937 \tab 472 \tab 179 \tab 27 \tab 41 \tab 73 \tab 131 \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 847 \tab 316 \tab 42 \tab 32 \tab 97 \tab 139 \tab 1473 \cr
   1985 \tab - \tab 884 \tab 319 \tab 51 \tab 34 \tab 100 \tab 146 \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 935 \tab 459 \tab 62 \tab 47 \tab 140 \tab 176 \tab 1819 \cr
   1988 \tab 504 \tab 496 \tab 249 \tab 27 \tab 31 \tab 75 \tab 99 \tab 1481 \cr
   1989 \tab 504 \tab 566 \tab 228 \tab 28 \tab 53 \tab 73 \tab 85 \tab 1537 \cr
   1990 \tab 455 \tab 486 \tab 211 \tab 27 \tab 37 \tab 58 \tab 98 \tab 1372 \cr
   1991 \tab 524 \tab 522 \tab 216 \tab 34 \tab 48 \tab 64 \tab 109 \tab 1517 \cr
   1993 \tab 531 \tab 605 \tab 219 \tab 37 \tab 38 \tab 65 \tab 111 \tab 1606 \cr
   1994 \tab 2508 \tab 275 \tab 94 \tab 10 \tab 21 \tab 39 \tab 45 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2992 \tab 717 \tab 363 \tab 74 \tab 20 \tab 158 \tab 186 \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1002 \tab 437 \tab 195 \tab 63 \tab 27 \tab 127 \tab 123 \tab 1974 \cr
   2014 \tab 1288 \tab 573 \tab 283 \tab 70 \tab 19 \tab 126 \tab 179 \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 49937 \tab 11387 \tab 5158 \tab 864 \tab 652 \tab 1879 \tab 2513 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} most impt
\item \code{2} second
\item \code{3} third
\item \code{4} fourth
\item \code{5} fifth
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
