% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_61.R
\name{jobinc}
\alias{jobinc}
\title{High income}
\source{
General Social Survey https://gss.norc.org
}
\description{
jobinc
}
\details{
Question 183a. Would you please look at this card and tell me which one thing on this list you would most prefer in a job? b. Which comes next? c. Which is third most important? d. Which is fourth most important? HIGH INCOME
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab fifth \tab fourth \tab most impt \tab no answer \tab second \tab third \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 102 \tab 286 \tab 272 \tab 42 \tab 352 \tab 450 \tab 1504 \cr
   1974 \tab - \tab 130 \tab 303 \tab 269 \tab 30 \tab 301 \tab 451 \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 78 \tab 281 \tab 295 \tab 50 \tab 326 \tab 469 \tab 1499 \cr
   1977 \tab - \tab 96 \tab 260 \tab 304 \tab 50 \tab 362 \tab 458 \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 67 \tab 235 \tab 288 \tab 32 \tab 383 \tab 463 \tab 1468 \cr
   1982 \tab 937 \tab 47 \tab 135 \tab 260 \tab 41 \tab 226 \tab 214 \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 61 \tab 272 \tab 276 \tab 32 \tab 357 \tab 475 \tab 1473 \cr
   1985 \tab - \tab 64 \tab 293 \tab 291 \tab 34 \tab 386 \tab 466 \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 91 \tab 254 \tab 436 \tab 47 \tab 466 \tab 525 \tab 1819 \cr
   1988 \tab 504 \tab 44 \tab 160 \tab 200 \tab 31 \tab 228 \tab 314 \tab 1481 \cr
   1989 \tab 504 \tab 37 \tab 174 \tab 210 \tab 53 \tab 245 \tab 314 \tab 1537 \cr
   1990 \tab 455 \tab 49 \tab 152 \tab 182 \tab 37 \tab 231 \tab 266 \tab 1372 \cr
   1991 \tab 524 \tab 51 \tab 169 \tab 228 \tab 48 \tab 233 \tab 264 \tab 1517 \cr
   1993 \tab 531 \tab 60 \tab 189 \tab 202 \tab 38 \tab 297 \tab 289 \tab 1606 \cr
   1994 \tab 2508 \tab 19 \tab 81 \tab 104 \tab 21 \tab 121 \tab 138 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2992 \tab 100 \tab 248 \tab 356 \tab 20 \tab 379 \tab 415 \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1002 \tab 64 \tab 144 \tab 257 \tab 25 \tab 261 \tab 221 \tab 1974 \cr
   2014 \tab 1288 \tab 71 \tab 167 \tab 337 \tab 16 \tab 344 \tab 315 \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 49937 \tab 1231 \tab 3803 \tab 4767 \tab 647 \tab 5498 \tab 6507 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} most impt
\item \code{2} second
\item \code{3} third
\item \code{4} fourth
\item \code{5} fifth
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
