% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_61.R
\name{joblose}
\alias{joblose}
\title{Is r likely to lose job}
\source{
General Social Survey https://gss.norc.org
}
\description{
joblose
}
\details{
Question 180a. Thinking about the next 12 months, how likely do you think it is that you will lose your job or be laid off?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab don't know \tab fairly likely \tab leaving labor force \tab no answer \tab not likely \tab not too likely \tab very likely \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 609 \tab 16 \tab 54 \tab 5 \tab 4 \tab 586 \tab 219 \tab 37 \tab - \tab 1530 \cr
   1978 \tab 638 \tab 9 \tab 33 \tab - \tab 5 \tab 629 \tab 182 \tab 36 \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 795 \tab 24 \tab 79 \tab - \tab 10 \tab 593 \tab 283 \tab 76 \tab - \tab 1860 \cr
   1983 \tab 654 \tab 15 \tab 72 \tab - \tab 5 \tab 564 \tab 233 \tab 56 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 583 \tab 10 \tab 42 \tab - \tab 9 \tab 609 \tab 218 \tab 63 \tab - \tab 1534 \cr
   1986 \tab 595 \tab 19 \tab 57 \tab - \tab 6 \tab 571 \tab 191 \tab 31 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 862 \tab 7 \tab 27 \tab - \tab 2 \tab 405 \tab 153 \tab 25 \tab - \tab 1481 \cr
   1989 \tab 921 \tab 5 \tab 23 \tab - \tab 3 \tab 424 \tab 136 \tab 25 \tab - \tab 1537 \cr
   1990 \tab 770 \tab 9 \tab 33 \tab - \tab 1 \tab 395 \tab 145 \tab 19 \tab - \tab 1372 \cr
   1991 \tab 907 \tab 6 \tab 40 \tab - \tab 1 \tab 374 \tab 152 \tab 37 \tab - \tab 1517 \cr
   1993 \tab 925 \tab 9 \tab 54 \tab - \tab 4 \tab 405 \tab 179 \tab 30 \tab - \tab 1606 \cr
   1994 \tab 1681 \tab 15 \tab 58 \tab - \tab 12 \tab 807 \tab 347 \tab 72 \tab - \tab 2992 \cr
   1996 \tab 1531 \tab 24 \tab 91 \tab - \tab 9 \tab 812 \tab 378 \tab 59 \tab - \tab 2904 \cr
   1998 \tab 1571 \tab 20 \tab 53 \tab - \tab 9 \tab 806 \tab 327 \tab 46 \tab - \tab 2832 \cr
   2000 \tab 1561 \tab 19 \tab 53 \tab - \tab 10 \tab 860 \tab 270 \tab 44 \tab - \tab 2817 \cr
   2002 \tab 2180 \tab 4 \tab 41 \tab - \tab 4 \tab 360 \tab 139 \tab 37 \tab - \tab 2765 \cr
   2004 \tab 1879 \tab 11 \tab 49 \tab - \tab 15 \tab 573 \tab 237 \tab 48 \tab - \tab 2812 \cr
   2006 \tab 3284 \tab 14 \tab 77 \tab - \tab 11 \tab 763 \tab 312 \tab 49 \tab - \tab 4510 \cr
   2008 \tab 1169 \tab 6 \tab 52 \tab - \tab 11 \tab 487 \tab 250 \tab 48 \tab - \tab 2023 \cr
   2010 \tab 1220 \tab 8 \tab 68 \tab - \tab 15 \tab 414 \tab 263 \tab 56 \tab - \tab 2044 \cr
   2012 \tab 1186 \tab 7 \tab 47 \tab - \tab 15 \tab 460 \tab 216 \tab 43 \tab - \tab 1974 \cr
   2014 \tab 1505 \tab 5 \tab 57 \tab - \tab 3 \tab 642 \tab 289 \tab 37 \tab - \tab 2538 \cr
   2016 \tab 1735 \tab 5 \tab 43 \tab - \tab 11 \tab 733 \tab 298 \tab 42 \tab - \tab 2867 \cr
   2018 \tab 1379 \tab 10 \tab 41 \tab - \tab 11 \tab 682 \tab 201 \tab 24 \tab - \tab 2348 \cr
   2021 \tab 2498 \tab - \tab 92 \tab - \tab 33 \tab 750 \tab 619 \tab 34 \tab 6 \tab 4032 \cr
   2022 \tab 2138 \tab 9 \tab 80 \tab - \tab 2 \tab 873 \tab 392 \tab 46 \tab 4 \tab 3544 \cr
   Total \tab 47126 \tab 286 \tab 1416 \tab 5 \tab 221 \tab 15577 \tab 6629 \tab 1120 \tab 10 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} fairly likely
\item \code{3} not too likely
\item \code{4} not likely
\item \code{5} leaving labor force
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
