% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_62.R
\name{jobmeans}
\alias{jobmeans}
\title{jobmeans}
\source{
General Social Survey https://gss.norc.org
}
\description{
Work important and feel accomplishment
}
\details{
Question 183a. Would you please look at this card and tell me which one thing on this list you would most prefer in a job? b. Which comes next? c. Which is third most important? d. Which is fourth most important? WORK IMPORTANT AND GIVES A FEELING OF ACCOMPLISHMENT
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab fifth \tab fourth \tab most impt \tab no answer \tab second \tab third \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 110 \tab 173 \tab 754 \tab 42 \tab 236 \tab 189 \tab 1504 \cr
   1974 \tab - \tab 103 \tab 148 \tab 742 \tab 30 \tab 279 \tab 182 \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 98 \tab 168 \tab 728 \tab 50 \tab 247 \tab 208 \tab 1499 \cr
   1977 \tab - \tab 123 \tab 181 \tab 695 \tab 50 \tab 287 \tab 194 \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 72 \tab 150 \tab 742 \tab 32 \tab 259 \tab 213 \tab 1468 \cr
   1982 \tab 937 \tab 99 \tab 144 \tab 346 \tab 41 \tab 152 \tab 141 \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 85 \tab 144 \tab 737 \tab 32 \tab 279 \tab 196 \tab 1473 \cr
   1985 \tab - \tab 78 \tab 171 \tab 715 \tab 34 \tab 300 \tab 236 \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 134 \tab 207 \tab 820 \tab 47 \tab 330 \tab 281 \tab 1819 \cr
   1988 \tab 504 \tab 62 \tab 106 \tab 466 \tab 31 \tab 190 \tab 122 \tab 1481 \cr
   1989 \tab 504 \tab 44 \tab 97 \tab 523 \tab 53 \tab 185 \tab 131 \tab 1537 \cr
   1990 \tab 455 \tab 43 \tab 102 \tab 460 \tab 37 \tab 173 \tab 102 \tab 1372 \cr
   1991 \tab 524 \tab 42 \tab 105 \tab 476 \tab 48 \tab 199 \tab 123 \tab 1517 \cr
   1993 \tab 531 \tab 48 \tab 118 \tab 530 \tab 38 \tab 171 \tab 170 \tab 1606 \cr
   1994 \tab 2508 \tab 30 \tab 59 \tab 228 \tab 21 \tab 83 \tab 63 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2992 \tab 110 \tab 219 \tab 615 \tab 20 \tab 297 \tab 257 \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1002 \tab 52 \tab 137 \tab 435 \tab 22 \tab 166 \tab 160 \tab 1974 \cr
   2014 \tab 1288 \tab 83 \tab 179 \tab 566 \tab 11 \tab 225 \tab 186 \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 49937 \tab 1416 \tab 2608 \tab 10578 \tab 639 \tab 4058 \tab 3154 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} most impt
\item \code{2} second
\item \code{3} third
\item \code{4} fourth
\item \code{5} fifth
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
