% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_61.R
\name{jobsec}
\alias{jobsec}
\title{jobsec}
\source{
General Social Survey https://gss.norc.org
}
\description{
No danger of being fired
}
\details{
Question 183a. Would you please look at this card and tell me which one thing on this list you would most prefer in a job? b. Which comes next? c. Which is third most important? d. Which is fourth most important? NO DANGER OF BEING FIRED
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab fifth \tab fourth \tab most impt \tab no answer \tab second \tab third \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 424 \tab 438 \tab 100 \tab 42 \tab 210 \tab 290 \tab 1504 \cr
   1974 \tab - \tab 481 \tab 417 \tab 111 \tab 30 \tab 159 \tab 286 \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 376 \tab 458 \tab 111 \tab 50 \tab 214 \tab 290 \tab 1499 \cr
   1977 \tab - \tab 362 \tab 467 \tab 119 \tab 50 \tab 208 \tab 324 \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 452 \tab 486 \tab 87 \tab 32 \tab 170 \tab 241 \tab 1468 \cr
   1982 \tab 937 \tab 169 \tab 242 \tab 93 \tab 41 \tab 175 \tab 203 \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 340 \tab 512 \tab 109 \tab 32 \tab 192 \tab 288 \tab 1473 \cr
   1985 \tab - \tab 387 \tab 528 \tab 109 \tab 34 \tab 168 \tab 308 \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 478 \tab 587 \tab 120 \tab 47 \tab 221 \tab 366 \tab 1819 \cr
   1988 \tab 504 \tab 264 \tab 299 \tab 62 \tab 31 \tab 131 \tab 190 \tab 1481 \cr
   1989 \tab 504 \tab 267 \tab 342 \tab 49 \tab 53 \tab 116 \tab 206 \tab 1537 \cr
   1990 \tab 455 \tab 236 \tab 288 \tab 55 \tab 37 \tab 106 \tab 195 \tab 1372 \cr
   1991 \tab 524 \tab 251 \tab 303 \tab 53 \tab 48 \tab 120 \tab 218 \tab 1517 \cr
   1993 \tab 531 \tab 242 \tab 348 \tab 91 \tab 38 \tab 146 \tab 210 \tab 1606 \cr
   1994 \tab 2508 \tab 98 \tab 152 \tab 39 \tab 21 \tab 71 \tab 103 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2992 \tab 412 \tab 403 \tab 170 \tab 20 \tab 225 \tab 288 \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1002 \tab 232 \tab 275 \tab 81 \tab 26 \tab 149 \tab 209 \tab 1974 \cr
   2014 \tab 1288 \tab 306 \tab 355 \tab 103 \tab 16 \tab 215 \tab 255 \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 49937 \tab 5777 \tab 6900 \tab 1662 \tab 648 \tab 2996 \tab 4470 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} most impt
\item \code{2} second
\item \code{3} third
\item \code{4} fourth
\item \code{5} fifth
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
