% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{letin1a}
\alias{letin1a}
\title{Number of immigrants nowadays should be}
\source{
General Social Survey https://gss.norc.org
}
\description{
letin1a
}
\details{
Question Do you think the number of immigrants to America nowadays should be...
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab don't know \tab increased a little \tab increased a lot \tab reduced a little \tab reduced a lot \tab remain the same as it is \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 16 \tab 52 \tab 30 \tab 222 \tab 256 \tab 330 \tab - \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 43 \tab 170 \tab 76 \tab 464 \tab 558 \tab 677 \tab 4 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 30 \tab 105 \tab 47 \tab 307 \tab 387 \tab 448 \tab 5 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 29 \tab 138 \tab 62 \tab 345 \tab 359 \tab 489 \tab 8 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 36 \tab 117 \tab 58 \tab 288 \tab 294 \tab 505 \tab 4 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 46 \tab 166 \tab 76 \tab 375 \tab 353 \tab 654 \tab 5 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 32 \tab 218 \tab 108 \tab 419 \tab 358 \tab 742 \tab 11 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 32 \tab 227 \tab 113 \tab 298 \tab 236 \tab 646 \tab 7 \tab - \tab 2348 \cr
   2021 \tab 1320 \tab 8 \tab 476 \tab 227 \tab 426 \tab 413 \tab 1128 \tab 3 \tab 31 \tab 4032 \cr
   2022 \tab - \tab 73 \tab 542 \tab 398 \tab 593 \tab 619 \tab 1288 \tab 17 \tab 14 \tab 3544 \cr
   Total \tab 54053 \tab 345 \tab 2211 \tab 1195 \tab 3737 \tab 3833 \tab 6907 \tab 64 \tab 45 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} increased a lot
\item \code{2} increased a little
\item \code{3} remain the same as it is
\item \code{4} reduced a little
\item \code{5} reduced a lot
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
