% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{libath}
\alias{libath}
\title{Allow anti-religious book in library}
\source{
General Social Survey https://gss.norc.org
}
\description{
libath
}
\details{
Question 76. There are always some people whose ideas are considered bad or dangerous by other people. For instance, somebody who is against churches and religion... c. If some people in your community suggested a book he wrote against churches and religion should be taken out of your public library, would you favor removing this book or not?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab don't know \tab no answer \tab not remove \tab remove \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 55 \tab 6 \tab 975 \tab 577 \tab - \tab - \tab 1613 \cr
   1973 \tab 30 \tab 3 \tab 915 \tab 556 \tab - \tab - \tab 1504 \cr
   1974 \tab 37 \tab 2 \tab 887 \tab 558 \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab - \tab - \tab - \tab 1490 \tab - \tab 1490 \cr
   1976 \tab 36 \tab 2 \tab 892 \tab 569 \tab - \tab - \tab 1499 \cr
   1977 \tab 27 \tab 5 \tab 894 \tab 604 \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab - \tab - \tab - \tab 1532 \tab - \tab 1532 \cr
   1980 \tab 38 \tab - \tab 909 \tab 521 \tab - \tab - \tab 1468 \cr
   1982 \tab 68 \tab 2 \tab 1071 \tab 719 \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab - \tab - \tab - \tab 1599 \tab - \tab 1599 \cr
   1984 \tab 36 \tab 3 \tab 937 \tab 497 \tab - \tab - \tab 1473 \cr
   1985 \tab 39 \tab 2 \tab 927 \tab 566 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab - \tab - \tab 1470 \tab - \tab 1470 \cr
   1987 \tab 43 \tab 10 \tab 1166 \tab 600 \tab - \tab - \tab 1819 \cr
   1988 \tab 20 \tab 5 \tab 622 \tab 330 \tab 504 \tab - \tab 1481 \cr
   1989 \tab 33 \tab 2 \tab 693 \tab 305 \tab 504 \tab - \tab 1537 \cr
   1990 \tab 31 \tab 5 \tab 609 \tab 272 \tab 455 \tab - \tab 1372 \cr
   1991 \tab 25 \tab 3 \tab 686 \tab 279 \tab 524 \tab - \tab 1517 \cr
   1993 \tab 41 \tab 6 \tab 721 \tab 307 \tab 531 \tab - \tab 1606 \cr
   1994 \tab 49 \tab 5 \tab 1384 \tab 558 \tab 996 \tab - \tab 2992 \cr
   1996 \tab 63 \tab 7 \tab 1307 \tab 546 \tab 981 \tab - \tab 2904 \cr
   1998 \tab 70 \tab 5 \tab 1295 \tab 512 \tab 950 \tab - \tab 2832 \cr
   2000 \tab 67 \tab 6 \tab 1262 \tab 526 \tab 956 \tab - \tab 2817 \cr
   2002 \tab 23 \tab 3 \tab 672 \tab 226 \tab 1841 \tab - \tab 2765 \cr
   2004 \tab 14 \tab 1 \tab 660 \tab 223 \tab 1914 \tab - \tab 2812 \cr
   2006 \tab 30 \tab 1 \tab 1452 \tab 520 \tab 2507 \tab - \tab 4510 \cr
   2008 \tab 34 \tab 3 \tab 962 \tab 353 \tab 671 \tab - \tab 2023 \cr
   2010 \tab 21 \tab 1 \tab 938 \tab 321 \tab 763 \tab - \tab 2044 \cr
   2012 \tab 31 \tab 2 \tab 973 \tab 302 \tab 666 \tab - \tab 1974 \cr
   2014 \tab 20 \tab 4 \tab 1312 \tab 378 \tab 824 \tab - \tab 2538 \cr
   2016 \tab 27 \tab 6 \tab 1436 \tab 421 \tab 977 \tab - \tab 2867 \cr
   2018 \tab 32 \tab 4 \tab 1211 \tab 327 \tab 774 \tab - \tab 2348 \cr
   2021 \tab - \tab 1 \tab 1106 \tab 204 \tab 2714 \tab 7 \tab 4032 \cr
   2022 \tab 18 \tab 2 \tab 917 \tab 222 \tab 2380 \tab 5 \tab 3544 \cr
   Total \tab 1058 \tab 107 \tab 29791 \tab 12899 \tab 28523 \tab 12 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} remove
\item \code{2} not remove
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
