% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{librac}
\alias{librac}
\title{librac}
\source{
General Social Survey https://gss.norc.org
}
\description{
Allow racists book in library
}
\details{
Question 78. Or consider a person who believes that Blacks are genetically inferior. c. If some people in your community suggested that a book he wrote which said Blacks are inferior should be taken out of your public library, would you favor removing this book, or not?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no answer \tab not remove \tab remove \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 55 \tab 7 \tab 892 \tab 545 \tab - \tab 1499 \cr
   1977 \tab - \tab 42 \tab 5 \tab 930 \tab 553 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 52 \tab 5 \tab 935 \tab 476 \tab - \tab 1468 \cr
   1982 \tab - \tab 86 \tab 3 \tab 1048 \tab 723 \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 47 \tab 3 \tab 926 \tab 497 \tab - \tab 1473 \cr
   1985 \tab - \tab 42 \tab 4 \tab 920 \tab 568 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 45 \tab 11 \tab 1115 \tab 648 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 30 \tab 2 \tab 603 \tab 342 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 32 \tab 6 \tab 669 \tab 326 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 37 \tab 4 \tab 587 \tab 289 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 26 \tab 4 \tab 650 \tab 313 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 42 \tab 3 \tab 691 \tab 339 \tab - \tab 1606 \cr
   1994 \tab 996 \tab 57 \tab 5 \tab 1313 \tab 621 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 66 \tab 6 \tab 1228 \tab 623 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 76 \tab 4 \tab 1184 \tab 618 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 62 \tab 11 \tab 1169 \tab 619 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 23 \tab 5 \tab 602 \tab 294 \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 13 \tab - \tab 594 \tab 291 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 33 \tab 2 \tab 1305 \tab 663 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 35 \tab 3 \tab 856 \tab 458 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 19 \tab 5 \tab 833 \tab 424 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 39 \tab 7 \tab 828 \tab 434 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 25 \tab 6 \tab 1077 \tab 606 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 34 \tab 6 \tab 1206 \tab 644 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 24 \tab 8 \tab 952 \tab 590 \tab - \tab 2348 \cr
   2021 \tab 2714 \tab 2 \tab 1 \tab 798 \tab 503 \tab 14 \tab 4032 \cr
   2022 \tab 2380 \tab 24 \tab 4 \tab 677 \tab 455 \tab 4 \tab 3544 \cr
   Total \tab 33124 \tab 1068 \tab 130 \tab 24588 \tab 13462 \tab 18 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} remove
\item \code{2} not remove
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
