% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_50.R
\name{life}
\alias{life}
\title{life}
\source{
General Social Survey https://gss.norc.org
}
\description{
Is life exciting or dull
}
\details{
Question 160. In general, do you find life exciting, pretty routine, or dull?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab don't know \tab dull \tab exciting \tab no answer \tab routine \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 5 \tab 76 \tab 675 \tab 15 \tab 733 \tab - \tab 1504 \cr
   1974 \tab - \tab 8 \tab 68 \tab 627 \tab 34 \tab 747 \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 9 \tab 54 \tab 660 \tab 15 \tab 761 \tab - \tab 1499 \cr
   1977 \tab - \tab 7 \tab 101 \tab 664 \tab 27 \tab 731 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 6 \tab 82 \tab 672 \tab - \tab 708 \tab - \tab 1468 \cr
   1982 \tab - \tab 9 \tab 122 \tab 795 \tab 8 \tab 926 \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 5 \tab 73 \tab 684 \tab 7 \tab 704 \tab - \tab 1473 \cr
   1985 \tab - \tab 13 \tab 98 \tab 727 \tab 4 \tab 692 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 10 \tab 81 \tab 778 \tab 39 \tab 911 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 7 \tab 47 \tab 428 \tab 20 \tab 475 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 3 \tab 54 \tab 453 \tab 13 \tab 510 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 5 \tab 45 \tab 408 \tab 5 \tab 454 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 8 \tab 41 \tab 434 \tab 5 \tab 505 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 4 \tab 69 \tab 496 \tab 4 \tab 502 \tab - \tab 1606 \cr
   1994 \tab 996 \tab 12 \tab 82 \tab 938 \tab 6 \tab 958 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 14 \tab 80 \tab 951 \tab 5 \tab 873 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 16 \tab 103 \tab 838 \tab 6 \tab 919 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 16 \tab 90 \tab 853 \tab 7 \tab 895 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 26 \tab 33 \tab 468 \tab - \tab 397 \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 6 \tab 37 \tab 451 \tab - \tab 404 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 28 \tab 102 \tab 984 \tab 1 \tab 888 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 14 \tab 64 \tab 647 \tab 1 \tab 626 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 11 \tab 66 \tab 639 \tab 4 \tab 561 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 11 \tab 62 \tab 672 \tab 1 \tab 562 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 14 \tab 82 \tab 855 \tab - \tab 763 \tab - \tab 2538 \cr
   2016 \tab 980 \tab 12 \tab 93 \tab 945 \tab 2 \tab 835 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 18 \tab 72 \tab 766 \tab 1 \tab 717 \tab - \tab 2348 \cr
   2021 \tab 1357 \tab 1 \tab 136 \tab 962 \tab - \tab 1571 \tab 5 \tab 4032 \cr
   2022 \tab 1203 \tab 4 \tab 114 \tab 993 \tab 1 \tab 1225 \tab 4 \tab 3544 \cr
   Total \tab 27605 \tab 302 \tab 2227 \tab 20463 \tab 231 \tab 21553 \tab 9 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} exciting
\item \code{2} routine
\item \code{3} dull
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
