% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_502.R
\name{localnum}
\alias{localnum}
\title{Number of employees: r's work site}
\source{
General Social Survey https://gss.norc.org
}
\description{
localnum
}
\details{
Question 1346. About how many people work at the location where you work? (Count part-time as well as full-time employees in all areas, departments and buildings.)
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab 1,000-1,999 \tab 1-9 \tab 10-49 \tab 100-499 \tab 2,000+ \tab 50-99 \tab 500-999 \tab don't know \tab no answer \tab skipped on web \tab 11 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 976 \tab 21 \tab 88 \tab 126 \tab 95 \tab 38 \tab 79 \tab 33 \tab 20 \tab 61 \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 605 \tab 26 \tab 231 \tab 244 \tab 166 \tab 60 \tab 112 \tab 48 \tab 13 \tab 12 \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1023 \tab 103 \tab 532 \tab 460 \tab 378 \tab 142 \tab 204 \tab 123 \tab 22 \tab 5 \tab - \tab - \tab 2992 \cr
   1996 \tab 901 \tab 73 \tab 539 \tab 443 \tab 390 \tab 161 \tab 236 \tab 123 \tab 30 \tab 8 \tab - \tab - \tab 2904 \cr
   1998 \tab 913 \tab 82 \tab 490 \tab 473 \tab 365 \tab 139 \tab 206 \tab 115 \tab 25 \tab 24 \tab - \tab - \tab 2832 \cr
   2000 \tab 926 \tab 95 \tab 490 \tab 442 \tab 361 \tab 126 \tab 216 \tab 119 \tab 24 \tab 18 \tab - \tab - \tab 2817 \cr
   2002 \tab 969 \tab 94 \tab 469 \tab 432 \tab 321 \tab 116 \tab 236 \tab 100 \tab 15 \tab 13 \tab - \tab - \tab 2765 \cr
   2004 \tab 945 \tab 63 \tab 502 \tab 444 \tab 353 \tab 142 \tab 216 \tab 100 \tab 10 \tab 37 \tab - \tab - \tab 2812 \cr
   2006 \tab 2656 \tab 74 \tab 524 \tab 378 \tab 354 \tab 144 \tab 238 \tab 108 \tab 22 \tab 12 \tab - \tab - \tab 4510 \cr
   2008 \tab 754 \tab 65 \tab 313 \tab 265 \tab 247 \tab 118 \tab 145 \tab 81 \tab 20 \tab 15 \tab - \tab - \tab 2023 \cr
   2010 \tab 857 \tab 39 \tab 325 \tab 271 \tab 212 \tab 93 \tab 156 \tab 58 \tab 15 \tab 18 \tab - \tab - \tab 2044 \cr
   2012 \tab 795 \tab 55 \tab 303 \tab 263 \tab 236 \tab 104 \tab 123 \tab 65 \tab 12 \tab 18 \tab - \tab - \tab 1974 \cr
   2014 \tab 993 \tab 59 \tab 396 \tab 354 \tab 306 \tab 135 \tab 184 \tab 93 \tab 12 \tab 6 \tab - \tab - \tab 2538 \cr
   2016 \tab 1141 \tab 62 \tab 432 \tab 407 \tab 318 \tab 174 \tab 195 \tab 98 \tab 19 \tab 21 \tab - \tab - \tab 2867 \cr
   2018 \tab 900 \tab 58 \tab 342 \tab 329 \tab 285 \tab 158 \tab 176 \tab 70 \tab 13 \tab 17 \tab - \tab - \tab 2348 \cr
   2021 \tab 1779 \tab 85 \tab 418 \tab 431 \tab 440 \tab 329 \tab 235 \tab 137 \tab 120 \tab 46 \tab 12 \tab - \tab 4032 \cr
   2022 \tab 1503 \tab 71 \tab 470 \tab 454 \tab 357 \tab 234 \tab 232 \tab 110 \tab - \tab 7 \tab 4 \tab 102 \tab 3544 \cr
   Total \tab 44970 \tab 1125 \tab 6864 \tab 6216 \tab 5184 \tab 2413 \tab 3189 \tab 1581 \tab 392 \tab 338 \tab 16 \tab 102 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} 1-9
\item \code{2} 10-49
\item \code{3} 50-99
\item \code{4} 100-499
\item \code{5} 500-999
\item \code{6} 1,000-1,999
\item \code{7} 2,000+
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
