% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_10.R
\name{madeg}
\alias{madeg}
\title{Mothers highest degree}
\source{
General Social Survey https://gss.norc.org
}
\description{
madeg
}
\details{
Question 21. If finished 9th-12th grade: Did she ever get a high school diploma or a GED certificate?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab associate/junior college \tab bachelor's \tab don't know \tab graduate \tab high school \tab iap \tab less than high school \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 16 \tab 39 \tab 96 \tab 13 \tab 390 \tab 80 \tab 832 \tab 147 \tab - \tab 1613 \cr
   1973 \tab 12 \tab 41 \tab 76 \tab 14 \tab 415 \tab 67 \tab 768 \tab 111 \tab - \tab 1504 \cr
   1974 \tab 13 \tab 44 \tab 98 \tab 16 \tab 420 \tab 55 \tab 817 \tab 21 \tab - \tab 1484 \cr
   1975 \tab 11 \tab 51 \tab 125 \tab 8 \tab 451 \tab 38 \tab 796 \tab 10 \tab - \tab 1490 \cr
   1976 \tab 13 \tab 45 \tab 136 \tab 9 \tab 422 \tab 51 \tab 811 \tab 12 \tab - \tab 1499 \cr
   1977 \tab 14 \tab 48 \tab 141 \tab 16 \tab 437 \tab 53 \tab 807 \tab 14 \tab - \tab 1530 \cr
   1978 \tab 27 \tab 49 \tab 92 \tab 11 \tab 512 \tab 51 \tab 781 \tab 9 \tab - \tab 1532 \cr
   1980 \tab 10 \tab 55 \tab 88 \tab 20 \tab 488 \tab 62 \tab 727 \tab 18 \tab - \tab 1468 \cr
   1982 \tab 21 \tab 82 \tab 181 \tab 21 \tab 579 \tab 66 \tab 889 \tab 21 \tab - \tab 1860 \cr
   1983 \tab 18 \tab 76 \tab 99 \tab 19 \tab 593 \tab 48 \tab 743 \tab 3 \tab - \tab 1599 \cr
   1984 \tab 20 \tab 57 \tab 85 \tab 16 \tab 564 \tab 68 \tab 659 \tab 4 \tab - \tab 1473 \cr
   1985 \tab 25 \tab 73 \tab 84 \tab 33 \tab 580 \tab 67 \tab 672 \tab - \tab - \tab 1534 \cr
   1986 \tab 20 \tab 73 \tab 101 \tab 14 \tab 570 \tab 47 \tab 640 \tab 5 \tab - \tab 1470 \cr
   1987 \tab 32 \tab 85 \tab 109 \tab 41 \tab 664 \tab 61 \tab 808 \tab 19 \tab - \tab 1819 \cr
   1988 \tab 28 \tab 70 \tab 83 \tab 19 \tab 594 \tab 37 \tab 647 \tab 3 \tab - \tab 1481 \cr
   1989 \tab 34 \tab 85 \tab 81 \tab 33 \tab 595 \tab 41 \tab 664 \tab 4 \tab - \tab 1537 \cr
   1990 \tab 32 \tab 80 \tab 69 \tab 22 \tab 562 \tab 32 \tab 556 \tab 19 \tab - \tab 1372 \cr
   1991 \tab 28 \tab 72 \tab 69 \tab 27 \tab 626 \tab 45 \tab 578 \tab 72 \tab - \tab 1517 \cr
   1993 \tab 43 \tab 103 \tab 100 \tab 30 \tab 697 \tab 41 \tab 571 \tab 21 \tab - \tab 1606 \cr
   1994 \tab 80 \tab 186 \tab 172 \tab 72 \tab 1299 \tab 135 \tab 1031 \tab 17 \tab - \tab 2992 \cr
   1996 \tab 94 \tab 213 \tab 188 \tab 84 \tab 1308 \tab 123 \tab 869 \tab 25 \tab - \tab 2904 \cr
   1998 \tab 109 \tab 232 \tab 175 \tab 86 \tab 1212 \tab 124 \tab 850 \tab 44 \tab - \tab 2832 \cr
   2000 \tab 104 \tab 219 \tab 168 \tab 79 \tab 1194 \tab 132 \tab 887 \tab 34 \tab - \tab 2817 \cr
   2002 \tab 119 \tab 210 \tab 103 \tab 80 \tab 1289 \tab 168 \tab 779 \tab 17 \tab - \tab 2765 \cr
   2004 \tab 138 \tab 270 \tab 120 \tab 112 \tab 1214 \tab 166 \tab 791 \tab 1 \tab - \tab 2812 \cr
   2006 \tab 139 \tab 263 \tab 78 \tab 122 \tab 1267 \tab 1687 \tab 924 \tab 30 \tab - \tab 4510 \cr
   2008 \tab 94 \tab 185 \tab 95 \tab 64 \tab 928 \tab 83 \tab 573 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 83 \tab 180 \tab 83 \tab 89 \tab 924 \tab 78 \tab 606 \tab 1 \tab - \tab 2044 \cr
   2012 \tab 110 \tab 189 \tab 83 \tab 97 \tab 866 \tab 65 \tab 563 \tab 1 \tab - \tab 1974 \cr
   2014 \tab 139 \tab 228 \tab 99 \tab 131 \tab 1160 \tab 83 \tab 698 \tab - \tab - \tab 2538 \cr
   2016 \tab 144 \tab 292 \tab 97 \tab 146 \tab 1349 \tab 101 \tab 736 \tab 2 \tab - \tab 2867 \cr
   2018 \tab 141 \tab 272 \tab 95 \tab 118 \tab 1048 \tab 84 \tab 589 \tab 1 \tab - \tab 2348 \cr
   2021 \tab 220 \tab 486 \tab 247 \tab 255 \tab 2019 \tab 100 \tab 698 \tab 1 \tab 6 \tab 4032 \cr
   2022 \tab 197 \tab 449 \tab - \tab 237 \tab 1562 \tab 422 \tab 677 \tab - \tab - \tab 3544 \cr
   Total \tab 2328 \tab 5102 \tab 3716 \tab 2154 \tab 28798 \tab 4561 \tab 25037 \tab 688 \tab 6 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} less than high school
\item \code{1} high school
\item \code{2} associate/junior college
\item \code{3} bachelor's
\item \code{4} graduate
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
