% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_512.R
\name{madenkid}
\alias{madenkid}
\title{If protestant, mothers denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
madenkid
}
\details{
Question 1386a. If Protestant, what specific denomination was that?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 7th day adventist \tab african methodist episcopal church \tab african methodist episcopal zion church \tab american baptist association \tab american lutheran church \tab assembly of god \tab baptist, don't know which \tab christian reform \tab christian scientist \tab christian; central christian \tab church of christ \tab church of god in christ \tab churches of god (except with christ and holiness) \tab congregationalist, 1st congreg \tab covenant \tab disciples of christ \tab don't know \tab dutch reform \tab episcopal church \tab evangelical united brethren \tab evangelical, evangelist \tab first christian \tab first christian disciples of christ \tab first reformed \tab free will baptist \tab full gospel \tab holiness; church of holiness \tab jehovah''s witnesses \tab lds \tab lds--jesus christ; church of jesus lds \tab lds--mormon \tab lds--reorganized \tab lutheran church - missouri synod \tab lutheran, don't know which \tab mennonite \tab methodist, don't know which \tab mission covenant \tab missionary baptist \tab moravian \tab mormon \tab nazarene \tab new birth christian \tab new testament christian \tab no answer \tab no denomination given or non-denom. church \tab other baptist churches \tab other fundamentalist \tab other lutheran churches \tab other methodist churches \tab pentecostal \tab pentecostal assembly of god \tab pilgrim holiness \tab presbyterian church u.s.a. \tab presbyterian, don't know which \tab primitive baptist \tab quaker \tab reformed \tab sanctified, sanctification \tab southern baptist convention \tab unitarian, universalist \tab united church of canada \tab united church of christ \tab united methodist church \tab united presbyterian church in the u.s.a. \tab wesleyan \tab wesleyan methodist--pilgrim \tab wisconsin evangelical lutheran synod \tab worldwide church of god \tab apostolic church \tab church of god in christ holiness \tab evangelical lutheran \tab evangelical reformed \tab independent \tab lutheran church in america \tab salvation army \tab american baptist churches in the u.s.a. \tab apostolic faith \tab brethren church, brethren \tab charismatic \tab christian & missionary alliance \tab ind bible, bible, bible fellowship \tab missionary church \tab national baptist convention of america \tab national baptist convention u.s.a., inc. \tab presbyterian church in the u.s. \tab united brethren, united brethren in christ \tab armenian apostolic church \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 707 \tab 4 \tab 3 \tab 1 \tab 2 \tab 6 \tab 7 \tab 182 \tab 2 \tab 4 \tab 18 \tab 14 \tab 2 \tab 10 \tab 8 \tab 1 \tab 4 \tab 13 \tab 1 \tab 23 \tab 2 \tab 4 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 3 \tab 15 \tab 2 \tab 1 \tab 1 \tab 4 \tab 72 \tab 3 \tab 99 \tab 1 \tab 3 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 120 \tab 12 \tab 3 \tab 1 \tab 2 \tab 1 \tab 17 \tab 1 \tab 1 \tab 2 \tab 44 \tab 1 \tab 2 \tab 1 \tab 1 \tab 34 \tab 1 \tab 1 \tab 3 \tab 19 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2126 \tab 12 \tab 27 \tab 1 \tab 1 \tab 1 \tab 5 \tab 159 \tab 1 \tab 7 \tab 16 \tab 9 \tab 2 \tab 6 \tab 8 \tab 1 \tab 2 \tab 20 \tab - \tab 10 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 3 \tab 10 \tab - \tab 1 \tab - \tab - \tab 3 \tab 49 \tab - \tab 93 \tab - \tab 1 \tab 1 \tab 8 \tab 2 \tab - \tab - \tab 83 \tab 16 \tab 5 \tab - \tab 2 \tab 3 \tab 8 \tab - \tab - \tab - \tab 38 \tab - \tab - \tab 1 \tab - \tab 42 \tab 1 \tab - \tab 4 \tab 23 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1254 \tab 3 \tab 5 \tab 3 \tab 15 \tab 10 \tab 3 \tab 135 \tab 3 \tab - \tab 27 \tab 15 \tab 1 \tab 6 \tab 5 \tab - \tab 4 \tab 31 \tab 1 \tab 30 \tab - \tab 2 \tab - \tab - \tab - \tab 6 \tab 1 \tab 5 \tab 9 \tab 8 \tab - \tab 4 \tab - \tab 11 \tab 30 \tab 2 \tab 30 \tab - \tab 3 \tab - \tab 11 \tab 3 \tab - \tab - \tab 13 \tab 41 \tab 2 \tab - \tab 1 \tab - \tab 19 \tab - \tab 1 \tab - \tab 37 \tab - \tab 3 \tab 4 \tab - \tab 103 \tab 1 \tab - \tab 2 \tab 66 \tab 7 \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 6 \tab - \tab 5 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab 6 \tab 3 \tab 6 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1748 \tab 7 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 186 \tab - \tab - \tab 55 \tab 13 \tab - \tab 3 \tab 4 \tab 1 \tab 1 \tab 27 \tab - \tab 21 \tab - \tab 5 \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab 12 \tab - \tab - \tab - \tab - \tab 4 \tab 42 \tab 3 \tab 58 \tab - \tab 3 \tab - \tab 14 \tab 1 \tab - \tab 1 \tab 3 \tab 24 \tab 1 \tab - \tab - \tab 1 \tab 21 \tab - \tab - \tab 1 \tab 27 \tab 1 \tab 3 \tab 2 \tab - \tab 15 \tab 2 \tab 1 \tab 2 \tab 12 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 69505 \tab 26 \tab 36 \tab 6 \tab 20 \tab 17 \tab 18 \tab 662 \tab 6 \tab 11 \tab 116 \tab 51 \tab 5 \tab 25 \tab 25 \tab 3 \tab 11 \tab 91 \tab 2 \tab 84 \tab 6 \tab 11 \tab 5 \tab 1 \tab 1 \tab 12 \tab 5 \tab 14 \tab 34 \tab 23 \tab 3 \tab 5 \tab 1 \tab 22 \tab 193 \tab 8 \tab 280 \tab 1 \tab 10 \tab 2 \tab 37 \tab 8 \tab 1 \tab 2 \tab 219 \tab 93 \tab 11 \tab 1 \tab 5 \tab 5 \tab 65 \tab 1 \tab 2 \tab 3 \tab 146 \tab 2 \tab 8 \tab 8 \tab 1 \tab 194 \tab 5 \tab 2 \tab 11 \tab 120 \tab 9 \tab 3 \tab 2 \tab 9 \tab 2 \tab 1 \tab 3 \tab 5 \tab 1 \tab 2 \tab 8 \tab 1 \tab 5 \tab 1 \tab 6 \tab 1 \tab 3 \tab 1 \tab 3 \tab 6 \tab 3 \tab 6 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{10000} american baptist association
\item \code{11000} american baptist churches in the u.s.a.
\item \code{12000} national baptist convention of america
\item \code{13000} national baptist convention u.s.a., inc.
\item \code{14000} southern baptist convention
\item \code{15000} other baptist churches
\item \code{18000} baptist, don't know which
\item \code{20000} african methodist episcopal church
\item \code{21000} african methodist episcopal zion church
\item \code{22000} united methodist church
\item \code{23000} other methodist churches
\item \code{28000} methodist, don't know which
\item \code{30000} american lutheran church
\item \code{31000} lutheran church in america
\item \code{32000} lutheran church - missouri synod
\item \code{33000} wisconsin evangelical lutheran synod
\item \code{34000} other lutheran churches
\item \code{35000} evangelical lutheran
\item \code{38000} lutheran, don't know which
\item \code{40000} presbyterian church in the u.s.
\item \code{41000} united presbyterian church in the u.s.a.
\item \code{42000} other presbyterian churches
\item \code{43000} presbyterian church u.s.a.
\item \code{48000} presbyterian, don't know which
\item \code{50000} episcopal church
\item \code{60001} hungarian reformed
\item \code{60002} evangelical congregational
\item \code{60003} ind bible, bible, bible fellowship
\item \code{60005} church of prophecy
\item \code{60006} new testament christian
\item \code{60007} church of god, saint & christ
\item \code{60008} moravian
\item \code{60009} christian & missionary alliance
\item \code{60010} advent christian
\item \code{60011} spiritualist
\item \code{60012} assembly of god
\item \code{60013} free methodist
\item \code{60014} apostolic faith
\item \code{60015} african methodist
\item \code{60016} free will baptist
\item \code{60017} eden evangelist
\item \code{60018} holiness (nazarene)
\item \code{60019} baptist (northern)
\item \code{60020} brethren church, brethren
\item \code{60021} witness holiness
\item \code{60022} brethren, plymouth
\item \code{60023} united brethren, united brethren in christ
\item \code{60024} independent
\item \code{60025} christian disciples
\item \code{60026} christ in christian union
\item \code{60027} open bible
\item \code{60028} christian catholic
\item \code{60029} christ church unity
\item \code{60030} christ adelphians
\item \code{60031} christian; central christian
\item \code{60032} christian reform
\item \code{60033} christian scientist
\item \code{60034} church of christ, evangelical
\item \code{60035} church of christ
\item \code{60036} churches of god (except with christ and holiness)
\item \code{60037} church of god in christ
\item \code{60038} church of god in christ holiness
\item \code{60039} church of the living god
\item \code{60040} congregationalist, 1st congreg
\item \code{60041} community church
\item \code{60042} covenant
\item \code{60043} dutch reform
\item \code{60044} disciples of christ
\item \code{60045} evangelical, evangelist
\item \code{60046} evangelical reformed
\item \code{60047} evangelist free church
\item \code{60048} first church
\item \code{60049} first christian disciples of christ
\item \code{60050} first reformed
\item \code{60051} first christian
\item \code{60052} full gospel
\item \code{60053} four square gospel
\item \code{60054} friends
\item \code{60055} holy roller
\item \code{60056} holiness; church of holiness
\item \code{60057} pilgrim holiness
\item \code{60058} jehovah''s witnesses
\item \code{60059} lds
\item \code{60060} lds--mormon
\item \code{60061} lds--reorganized
\item \code{60062} lds--jesus christ; church of jesus lds
\item \code{60063} mennonite
\item \code{60064} mormon
\item \code{60065} nazarene
\item \code{60066} pentecostal assembly of god
\item \code{60067} pentecostal church of god
\item \code{60068} pentecostal
\item \code{60069} pentecostal holiness, holiness pentecostal
\item \code{60070} quaker
\item \code{60071} reformed
\item \code{60072} reformed united church of christ
\item \code{60073} reformed church of christ
\item \code{60074} religious science
\item \code{60075} mind science
\item \code{60076} salvation army
\item \code{60077} 7th day adventist
\item \code{60078} sanctified, sanctification
\item \code{60079} united holiness
\item \code{60080} unitarian, universalist
\item \code{60081} united church of christ
\item \code{60082} united church, unity church
\item \code{60083} wesleyan
\item \code{60084} wesleyan methodist--pilgrim
\item \code{60085} zion union
\item \code{60086} zion union apostolic
\item \code{60087} zion union apostolic--reformed
\item \code{60088} disciples of god
\item \code{60089} grace reformed
\item \code{60090} holiness church of god
\item \code{60091} evangelical covenant
\item \code{60092} mission covenant
\item \code{60093} missionary baptist
\item \code{60094} swedish mission
\item \code{60095} unity
\item \code{60096} united church of christianity
\item \code{60097} other fundamentalist
\item \code{60098} federated church
\item \code{60099} american reform
\item \code{60100} grace brethren
\item \code{60101} christ in god
\item \code{60102} charismatic
\item \code{60103} pentecostal apostolic
\item \code{60104} house of prayer
\item \code{60105} latvian lutheran
\item \code{60106} triumph church of god
\item \code{60107} apostolic christian
\item \code{60108} christ cathedral of truth
\item \code{60109} bible missionary
\item \code{60110} calvary bible
\item \code{60111} amish
\item \code{60112} evangelical methodist
\item \code{60113} worldwide church of god
\item \code{60114} church universal and triumphant
\item \code{60115} mennonite brethren
\item \code{60116} church of the first born
\item \code{60117} missionary church
\item \code{60118} the way ministry
\item \code{60119} united church of canada
\item \code{60120} evangelical united brethren
\item \code{60121} the church of god of prophecy
\item \code{60122} chapel of faith
\item \code{60123} polish national church
\item \code{60124} faith gospel tabernacle
\item \code{60125} christian calvary chapel
\item \code{60126} carmelite
\item \code{60127} church of daniel's band
\item \code{60128} christian tabernacle
\item \code{60129} living word
\item \code{60130} true light church of christ
\item \code{60131} macedonia
\item \code{60132} brother of christ
\item \code{60133} primitive baptist
\item \code{60134} independent fundamental church of america
\item \code{60135} chinese gospel church
\item \code{60136} new age spirituality
\item \code{60137} new song
\item \code{60138} apostolic church
\item \code{60139} faith christian
\item \code{60140} people's church
\item \code{60141} new birth christian
\item \code{60142} unity school of christianity
\item \code{60143} assyrian evangelist church
\item \code{60144} spirit of christ
\item \code{60145} church of jesus christ of the restoration
\item \code{60146} laotian christian
\item \code{60148} schwenkfelder
\item \code{60149} polish catholic
\item \code{60150} zwinglian
\item \code{60151} world overcomer outreach ministry
\item \code{60152} course in miracles
\item \code{60153} unity of the brethren
\item \code{60154} spirit filled
\item \code{60155} christian union
\item \code{60156} church of living christ
\item \code{60157} community of christ
\item \code{60158} new hope christian fellowship
\item \code{60159} community christian fellowship
\item \code{60160} friends in christ
\item \code{60161} hawaiian ohana
\item \code{60162} reformed church of jesus christ of latter day saints
\item \code{60163} swedenurgian/churches of the new jerusalem
\item \code{60164} divine science
\item \code{60166} united christian
\item \code{60167} sanctuary
\item \code{60168} rain on us deliverance ministries
\item \code{60169} the word church
\item \code{60170} cornerstone church
\item \code{60171} life sanctuary
\item \code{60172} word of faith church
\item \code{60173} harvest church
\item \code{60174} shephard''s chapel
\item \code{60175} greater new testament church
\item \code{60176} vineyard church
\item \code{60177} real life ministries
\item \code{60178} cathedral of joy
\item \code{60179} great faith ministries
\item \code{60180} shield of faith ministries
\item \code{60181} born again
\item \code{60182} alliance
\item \code{60183} jacobite apostolic
\item \code{60184} church of god of israel
\item \code{60185} journeys
\item \code{60186} national progressive baptist
\item \code{60187} new apostolic
\item \code{60188} metropolitan community
\item \code{60189} family life church
\item \code{60190} faith fellowship
\item \code{60191} faith covenant
\item \code{60192} new thought
\item \code{60193} free spirit ministry
\item \code{60194} jacobite syrian christian church
\item \code{60195} the ark church
\item \code{60196} empowerment temple
\item \code{60197} grace independent baptist church
\item \code{60198} new life
\item \code{60199} follower of christ
\item \code{60200} north country chapel
\item \code{60201} pathways christian church
\item \code{60202} messianic christian
\item \code{60203} warehouse ministries
\item \code{60204} universal church of the kingdom of god
\item \code{60205} assembly of christ
\item \code{60206} the amana church
\item \code{60207} the legacy church
\item \code{60208} calvary
\item \code{60209} church of the living god
\item \code{60210} ethiopian evangelical church
\item \code{60211} disciple of jesus
\item \code{60212} scandinavian church
\item \code{60213} hebrew roots
\item \code{60214} hebrew isrealites
\item \code{60215} armenian apostolic church
\item \code{60216} word of jesus
\item \code{60217} old testament
\item \code{70000} no denomination given or non-denom. church
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
