% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_10.R
\name{maeduc}
\alias{maeduc}
\title{maeduc}
\source{
General Social Survey https://gss.norc.org
}
\description{
Highest year school completed, mother
}
\details{
Question 17. What is the highest grade in elementary school or high school that your mother finished and got credit for?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab iap \tab no answer \tab no formal schooling \tab 20 \tab Total \cr
   1972 \tab 8 \tab 85 \tab 54 \tab 328 \tab 19 \tab 60 \tab 14 \tab 42 \tab 3 \tab 7 \tab 3 \tab 12 \tab 43 \tab 43 \tab 43 \tab 78 \tab 47 \tab 262 \tab 36 \tab 263 \tab 80 \tab 16 \tab 67 \tab - \tab 1613 \cr
   1973 \tab 7 \tab 74 \tab 60 \tab 342 \tab 28 \tab 50 \tab 8 \tab 47 \tab 2 \tab 4 \tab - \tab 7 \tab 32 \tab 38 \tab 41 \tab 76 \tab 39 \tab 254 \tab 40 \tab 207 \tab 67 \tab 22 \tab 56 \tab 3 \tab 1504 \cr
   1974 \tab 2 \tab 90 \tab 61 \tab 331 \tab 33 \tab 53 \tab 10 \tab 48 \tab 3 \tab 10 \tab 2 \tab 10 \tab 25 \tab 38 \tab 48 \tab 78 \tab 35 \tab 243 \tab 47 \tab 201 \tab 55 \tab 1 \tab 58 \tab 2 \tab 1484 \cr
   1975 \tab 4 \tab 90 \tab 57 \tab 360 \tab 33 \tab 49 \tab 7 \tab 53 \tab 1 \tab 8 \tab 2 \tab 7 \tab 30 \tab 34 \tab 40 \tab 64 \tab 71 \tab 253 \tab 53 \tab 193 \tab 38 \tab 3 \tab 40 \tab - \tab 1490 \cr
   1976 \tab 4 \tab 66 \tab 67 \tab 354 \tab 25 \tab 46 \tab 11 \tab 47 \tab 2 \tab 2 \tab 1 \tab 12 \tab 35 \tab 35 \tab 27 \tab 80 \tab 54 \tab 243 \tab 41 \tab 236 \tab 51 \tab 6 \tab 50 \tab 4 \tab 1499 \cr
   1977 \tab 1 \tab 78 \tab 59 \tab 378 \tab 23 \tab 44 \tab 10 \tab 50 \tab 3 \tab 9 \tab 1 \tab 1 \tab 21 \tab 49 \tab 27 \tab 63 \tab 57 \tab 256 \tab 44 \tab 243 \tab 53 \tab 4 \tab 56 \tab - \tab 1530 \cr
   1978 \tab 1 \tab 71 \tab 67 \tab 421 \tab 22 \tab 65 \tab 16 \tab 48 \tab 5 \tab 7 \tab 1 \tab 8 \tab 18 \tab 31 \tab 26 \tab 74 \tab 42 \tab 242 \tab 49 \tab 216 \tab 51 \tab 5 \tab 41 \tab 5 \tab 1532 \cr
   1980 \tab 5 \tab 67 \tab 46 \tab 383 \tab 34 \tab 69 \tab 5 \tab 62 \tab 7 \tab 9 \tab - \tab 4 \tab 27 \tab 41 \tab 29 \tab 75 \tab 39 \tab 227 \tab 50 \tab 181 \tab 62 \tab 3 \tab 41 \tab 2 \tab 1468 \cr
   1982 \tab 4 \tab 96 \tab 60 \tab 491 \tab 39 \tab 59 \tab 22 \tab 79 \tab 8 \tab 10 \tab - \tab 11 \tab 30 \tab 23 \tab 49 \tab 88 \tab 42 \tab 229 \tab 62 \tab 313 \tab 66 \tab 13 \tab 62 \tab 4 \tab 1860 \cr
   1983 \tab 2 \tab 70 \tab 68 \tab 483 \tab 40 \tab 73 \tab 14 \tab 76 \tab 8 \tab 9 \tab 4 \tab 13 \tab 24 \tab 33 \tab 33 \tab 61 \tab 46 \tab 221 \tab 46 \tab 179 \tab 48 \tab 3 \tab 44 \tab 1 \tab 1599 \cr
   1984 \tab 1 \tab 103 \tab 50 \tab 450 \tab 35 \tab 68 \tab 21 \tab 57 \tab 6 \tab 7 \tab 2 \tab 7 \tab 19 \tab 21 \tab 28 \tab 43 \tab 32 \tab 201 \tab 55 \tab 161 \tab 68 \tab 4 \tab 33 \tab 1 \tab 1473 \cr
   1985 \tab - \tab 74 \tab 59 \tab 482 \tab 36 \tab 62 \tab 11 \tab 73 \tab 7 \tab 18 \tab 6 \tab 12 \tab 25 \tab 26 \tab 27 \tab 60 \tab 35 \tab 201 \tab 48 \tab 172 \tab 67 \tab - \tab 31 \tab 2 \tab 1534 \cr
   1986 \tab - \tab 59 \tab 48 \tab 472 \tab 25 \tab 75 \tab 10 \tab 71 \tab 7 \tab 14 \tab 2 \tab 5 \tab 24 \tab 22 \tab 29 \tab 50 \tab 40 \tab 193 \tab 48 \tab 191 \tab 47 \tab 4 \tab 33 \tab 1 \tab 1470 \cr
   1987 \tab 2 \tab 92 \tab 73 \tab 508 \tab 51 \tab 103 \tab 17 \tab 87 \tab 9 \tab 25 \tab 1 \tab 9 \tab 29 \tab 29 \tab 35 \tab 76 \tab 33 \tab 238 \tab 71 \tab 205 \tab 61 \tab 13 \tab 42 \tab 10 \tab 1819 \cr
   1988 \tab 2 \tab 66 \tab 48 \tab 470 \tab 56 \tab 79 \tab 15 \tab 74 \tab 4 \tab 8 \tab 3 \tab 6 \tab 28 \tab 28 \tab 23 \tab 48 \tab 30 \tab 176 \tab 43 \tab 177 \tab 37 \tab 18 \tab 36 \tab 6 \tab 1481 \cr
   1989 \tab 2 \tab 87 \tab 36 \tab 475 \tab 47 \tab 78 \tab 26 \tab 92 \tab 4 \tab 20 \tab 2 \tab 7 \tab 23 \tab 19 \tab 28 \tab 63 \tab 29 \tab 182 \tab 49 \tab 162 \tab 41 \tab 2 \tab 60 \tab 3 \tab 1537 \cr
   1990 \tab 1 \tab 54 \tab 50 \tab 436 \tab 54 \tab 71 \tab 22 \tab 80 \tab 6 \tab 14 \tab 1 \tab 3 \tab 24 \tab 16 \tab 26 \tab 38 \tab 19 \tab 160 \tab 52 \tab 151 \tab 32 \tab 33 \tab 25 \tab 4 \tab 1372 \cr
   1991 \tab 1 \tab 71 \tab 43 \tab 455 \tab 56 \tab 84 \tab 16 \tab 80 \tab 8 \tab 18 \tab - \tab 7 \tab 15 \tab 14 \tab 24 \tab 39 \tab 32 \tab 192 \tab 49 \tab 176 \tab 45 \tab 63 \tab 24 \tab 5 \tab 1517 \cr
   1993 \tab 1 \tab 76 \tab 47 \tab 573 \tab 47 \tab 101 \tab 22 \tab 102 \tab 5 \tab 17 \tab 3 \tab 8 \tab 14 \tab 23 \tab 23 \tab 40 \tab 37 \tab 177 \tab 35 \tab 159 \tab 41 \tab 27 \tab 23 \tab 5 \tab 1606 \cr
   1994 \tab 3 \tab 141 \tab 100 \tab 1007 \tab 95 \tab 197 \tab 47 \tab 187 \tab 15 \tab 38 \tab 7 \tab 16 \tab 29 \tab 38 \tab 33 \tab 87 \tab 58 \tab 319 \tab 86 \tab 284 \tab 135 \tab 13 \tab 41 \tab 16 \tab 2992 \cr
   1996 \tab 3 \tab 130 \tab 93 \tab 1009 \tab 76 \tab 243 \tab 46 \tab 214 \tab 16 \tab 48 \tab 8 \tab 8 \tab 23 \tab 31 \tab 31 \tab 69 \tab 48 \tab 243 \tab 75 \tab 292 \tab 123 \tab 17 \tab 34 \tab 24 \tab 2904 \cr
   1998 \tab 3 \tab 129 \tab 103 \tab 932 \tab 98 \tab 204 \tab 44 \tab 235 \tab 22 \tab 47 \tab 5 \tab 16 \tab 26 \tab 27 \tab 33 \tab 87 \tab 36 \tab 229 \tab 72 \tab 269 \tab 124 \tab 40 \tab 32 \tab 19 \tab 2832 \cr
   2000 \tab - \tab 129 \tab 81 \tab 939 \tab 88 \tab 201 \tab 41 \tab 218 \tab 17 \tab 41 \tab 4 \tab 17 \tab 39 \tab 34 \tab 23 \tab 79 \tab 37 \tab 233 \tab 77 \tab 284 \tab 132 \tab 33 \tab 46 \tab 24 \tab 2817 \cr
   2002 \tab 3 \tab 121 \tab 84 \tab 997 \tab 116 \tab 205 \tab 40 \tab 206 \tab 12 \tab 52 \tab 6 \tab 14 \tab 28 \tab 27 \tab 23 \tab 71 \tab 36 \tab 224 \tab 68 \tab 178 \tab 168 \tab 21 \tab 44 \tab 21 \tab 2765 \cr
   2004 \tab 1 \tab 116 \tab 115 \tab 945 \tab 82 \tab 236 \tab 55 \tab 282 \tab 14 \tab 48 \tab 12 \tab 13 \tab 25 \tab 14 \tab 23 \tab 70 \tab 40 \tab 179 \tab 86 \tab 181 \tab 166 \tab 36 \tab 39 \tab 34 \tab 2812 \cr
   2006 \tab 7 \tab 114 \tab 93 \tab 971 \tab 117 \tab 233 \tab 42 \tab 261 \tab 18 \tab 69 \tab 14 \tab 22 \tab 37 \tab 32 \tab 29 \tab 109 \tab 36 \tab 222 \tab 78 \tab 167 \tab 1687 \tab 31 \tab 94 \tab 27 \tab 4510 \cr
   2008 \tab 5 \tab 100 \tab 65 \tab 725 \tab 73 \tab 143 \tab 27 \tab 182 \tab 12 \tab 50 \tab 1 \tab 7 \tab 19 \tab 17 \tab 23 \tab 69 \tab 25 \tab 128 \tab 43 \tab 156 \tab 83 \tab 4 \tab 52 \tab 14 \tab 2023 \cr
   2010 \tab 4 \tab 87 \tab 68 \tab 716 \tab 72 \tab 144 \tab 29 \tab 191 \tab 20 \tab 46 \tab 8 \tab 9 \tab 21 \tab 24 \tab 20 \tab 58 \tab 24 \tab 153 \tab 47 \tab 160 \tab 78 \tab 5 \tab 41 \tab 19 \tab 2044 \cr
   2012 \tab 3 \tab 68 \tab 71 \tab 680 \tab 68 \tab 158 \tab 36 \tab 185 \tab 12 \tab 51 \tab 8 \tab 6 \tab 26 \tab 15 \tab 21 \tab 73 \tab 24 \tab 113 \tab 50 \tab 148 \tab 74 \tab 7 \tab 49 \tab 28 \tab 1974 \cr
   2014 \tab 3 \tab 101 \tab 84 \tab 902 \tab 76 \tab 223 \tab 38 \tab 230 \tab 18 \tab 77 \tab 15 \tab 10 \tab 32 \tab 31 \tab 29 \tab 64 \tab 27 \tab 146 \tab 63 \tab 181 \tab 83 \tab 3 \tab 68 \tab 34 \tab 2538 \cr
   2016 \tab 5 \tab 98 \tab 86 \tab 1055 \tab 100 \tab 236 \tab 53 \tab 302 \tab 20 \tab 89 \tab 10 \tab 11 \tab 27 \tab 16 \tab 32 \tab 77 \tab 29 \tab 182 \tab 61 \tab 180 \tab 101 \tab 5 \tab 59 \tab 33 \tab 2867 \cr
   2018 \tab 6 \tab 70 \tab 69 \tab 826 \tab 71 \tab 208 \tab 43 \tab 278 \tab 24 \tab 59 \tab 7 \tab 10 \tab 26 \tab 19 \tab 33 \tab 67 \tab 28 \tab 98 \tab 65 \tab 169 \tab 84 \tab 6 \tab 52 \tab 30 \tab 2348 \cr
   2021 \tab 4 \tab 146 \tab 79 \tab 1580 \tab 163 \tab 374 \tab 84 \tab 517 \tab 48 \tab 106 \tab 15 \tab 7 \tab 15 \tab 17 \tab 30 \tab 57 \tab 34 \tab 144 \tab 71 \tab 316 \tab 100 \tab 3 \tab 55 \tab 67 \tab 4032 \cr
   2022 \tab 2 \tab 118 \tab 87 \tab 1340 \tab 120 \tab 272 \tab 55 \tab 470 \tab 40 \tab 109 \tab 11 \tab 7 \tab 31 \tab 23 \tab 13 \tab 81 \tab 27 \tab 120 \tab 68 \tab - \tab 422 \tab - \tab 50 \tab 78 \tab 3544 \cr
   Total \tab 100 \tab 3137 \tab 2331 \tab 22816 \tab 2118 \tab 4566 \tab 957 \tab 5226 \tab 406 \tab 1146 \tab 165 \tab 322 \tab 890 \tab 928 \tab 1002 \tab 2312 \tab 1268 \tab 6883 \tab 1928 \tab 6751 \tab 4570 \tab 464 \tab 1578 \tab 526 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} no formal schooling
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
