% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_09.R
\name{maind10}
\alias{maind10}
\title{maind10}
\source{
General Social Survey https://gss.norc.org
}
\description{
Mothers industry code (naics 2007)
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab accounting, tax preparation, bookkeeping, and payroll services \tab administration of economic programs and space research \tab administration of environmental quality and housing programs \tab administration of human resource programs \tab aerospace product and parts manufacturing \tab agricultural implement manufacturing \tab air transportation \tab aircraft and parts manufacturing \tab animal food, grain, and oilseed milling \tab animal production \tab animal slaughtering and processing \tab apparel accessories and other apparel manufacturing \tab apparel, fabrics, and notions, merchant wholesalers \tab architectural, engineering, and related services \tab automobile dealers \tab automotive equipment rental and leasing \tab automotive repair and maintenance \tab bakeries, except retail \tab banking and related activities \tab beauty salons \tab beer, wine, and liquor stores \tab beverage manufacturing \tab building material and supplies dealers \tab bus service and urban transit \tab business support services \tab business, professional, political, and similar organizations \tab carpet and rug mills \tab child day care services \tab civic, social, advocacy organizations, and grantmaking and giving services \tab clothing stores \tab coal mining \tab colleges and universities, including junior colleges \tab commercial and industrial machinery and equipment repair and maintenance \tab commercial and service industry machinery manufacturing \tab communications, and audio and video equipment manufacturing \tab community food and housing, and emergency services \tab computer and peripheral equipment manufacturing \tab computer systems design and related services \tab construction \tab crop production \tab cut and sew apparel manufacturing \tab dairy product manufacturing \tab department stores and discount stores \tab don't know \tab drinking places, alcoholic beverages \tab drycleaning and laundry services \tab electric power generation, transmission, and distribution \tab electrical and electronic goods, merchant wholesalers \tab electrical lighting and electrical equipment manufacturing, and other electrical component manufacturing, n.e.c. \tab electronic component and product manufacturing, n.e.c. \tab elementary and secondary schools \tab employment services \tab executive offices and legislative bodies \tab fabric mills, except knitting mills \tab farm product raw materials, merchant wholesalers \tab farm supplies, merchant wholesalers \tab fiber, yarn, and thread mills \tab footwear manufacturing \tab forestry, except logging \tab foundries \tab fruit and vegetable preserving and specialty food manufacturing \tab funeral homes, and cemeteries and crematories \tab furniture and home furnishings stores \tab furniture and related product manufacturing \tab gasoline stations \tab gift, novelty, and souvenir shops \tab glass and glass product manufacturing \tab groceries and related products, merchant wholesalers \tab grocery stores \tab hardware stores \tab health and personal care, except drug, stores \tab home health care services \tab hospitals \tab household appliance manufacturing \tab household appliance stores \tab independent artists, performing arts, spectator sports, and related industries \tab individual and family services \tab insurance carriers and related activities \tab investigation and security services \tab iron and steel mills and steel product manufacturing \tab jewelry, luggage, and leather goods stores \tab justice, public order, and safety activities \tab knitting fabric mills, and apparel knitting mills \tab landscaping services \tab leather tanning and finishing and other allied products manufacturing \tab legal services \tab libraries and archives \tab logging \tab machine shops; turned product; screw, nut, and bolt manufacturing \tab machinery manufacturing, n.e.c. \tab mail order houses \tab management of companies and enterprises \tab metal forgings and stampings \tab metalworking machinery manufacturing \tab miscellaneous durable goods, merchant wholesalers \tab miscellaneous fabricated metal products manufacturing \tab miscellaneous general merchandise stores \tab miscellaneous manufacturing, n.e.c. \tab miscellaneous paper and pulp products \tab miscellaneous retail stores \tab miscellaneous wood products \tab motion pictures and video industries \tab motor vehicles and motor vehicle equipment manufacturing \tab motor vehicles, parts and supplies, merchant wholesalers \tab museums, art galleries, historical sites, and similar institutions \tab nail salons and other personal care services \tab national security and international affairs \tab natural gas distribution \tab navigational, measuring, electromedical, and control instruments manufacturing \tab newspaper publishers \tab no answer \tab non-depository credit and related activities \tab not specified food industries \tab not specified machinery manufacturing \tab not specified manufacturing industries \tab not specified retail trade \tab not specified utilities \tab not specified wholesale trade \tab nursing care facilities \tab office supplies and stationery stores \tab offices of dentists \tab offices of other health practitioners \tab offices of physicians \tab oil and gas extraction \tab ordnance \tab other administrative and other support services \tab other amusement, gambling, and recreation industries \tab other direct selling establishments \tab other health care services \tab other personal services \tab other professional, scientific, and technical services \tab other schools and instruction, and educational support services \tab outpatient care centers \tab paper and paper products, merchant wholesalers \tab paperboard containers and boxes \tab periodical, book, and directory publishers \tab personal and household goods repair and maintenance \tab petroleum and petroleum products, merchant wholesalers \tab petroleum refining \tab pharmaceutical and medicine manufacturing \tab pharmacies and drug stores \tab plastics product manufacturing \tab postal service \tab prefabricated wood buildings and mobile homes \tab printing and related support activities \tab private households \tab professional and commercial equipment and supplies, merchant wholesalers \tab radio and television broadcasting and cable subscription programming \tab radio, tv, and computer stores \tab rail transportation \tab real estate \tab recreational vehicle parks and camps, and rooming and boarding houses \tab religious organizations \tab residential care facilities, without nursing \tab resin, synthetic rubber and fibers, and filaments manufacturing \tab restaurants and other food services \tab retail bakeries \tab retail florists \tab savings institutions, including credit unions \tab scenic and sightseeing transportation \tab scientific research and development services \tab seafood and other miscellaneous foods, n.e.c. \tab securities, commodities, funds, trusts, and other financial investments \tab services incidental to transportation \tab services to buildings and dwellings \tab sewing, needlework, and piece goods stores \tab ship and boat building \tab shoe stores \tab soap, cleaning compound, and cosmetics manufacturing \tab specialized design services \tab specialty food stores \tab sporting and athletic goods, and doll, toy and game manufacturing \tab sporting goods, camera, and hobby and toy stores \tab structural metals, and boiler, tank, and shipping container manufacturing \tab sugar and confectionery products \tab support activities for agriculture and forestry \tab taxi and limousine service \tab textile and fabric finishing and coating mills \tab textile product mills, except carpet and rug \tab tire manufacturing \tab tobacco manufacturing \tab travel arrangements and reservation services \tab traveler accommodation \tab truck transportation \tab u.s. air force \tab u.s. army \tab uncodeable \tab used merchandise stores \tab vending machine operators \tab veterinary services \tab warehousing and storage \tab waste management and remediation services \tab water transportation \tab wired telecommunications carriers \tab advertising and related services \tab agricultural chemical manufacturing \tab alcoholic beverages, merchant wholesalers \tab auto parts, accessories, and tire stores \tab bowling centers \tab business, technical, and trade schools and training \tab cutlery and hand tool manufacturing \tab electronic and precision equipment repair and maintenance \tab fuel dealers \tab hardware, plumbing and heating equipment, and supplies, merchant wholesalers \tab industrial and miscellaneous chemicals \tab labor unions \tab lawn and garden equipment and supplies stores \tab machinery, equipment, and supplies, merchant wholesalers \tab management, scientific, and technical consulting services \tab medical equipment and supplies manufacturing \tab metals and minerals, except petroleum, merchant wholesalers \tab miscellaneous nondurable goods, merchant wholesalers \tab music stores \tab offices of optometrists \tab other general government and support \tab public finance activities \tab pulp, paper, and paperboard mills \tab recyclable material, merchant wholesalers \tab rubber product, except tire, manufacturing \tab structural clay product manufacturing \tab vocational rehabilitation services \tab water, steam, air-conditioning, and irrigation systems \tab wholesale electronic markets, agents and brokers \tab book stores and news dealers \tab drugs, sundries, and chemical and allied products, merchant wholesalers \tab electric and gas, and other combinations \tab metal ore mining \tab not specified metal industries \tab offices of chiropractors \tab other motor vehicle dealers \tab paint, coating, and adhesive manufacturing \tab sound recording industries \tab u.s. marine corps \tab veneer, plywood, and engineered wood products \tab couriers and messengers \tab furniture and home furnishing, merchant wholesalers \tab other telecommunications services \tab pottery, ceramics, and plumbing fixture manufacturing \tab sawmills and wood preservation \tab support activities for mining \tab video tape and disk rental \tab cement, concrete, lime, and gypsum product manufacturing \tab data processing, hosting, and related services \tab engines, turbines, and power transmission equipment manufacturing \tab u.s. armed forces/n.s. \tab u.s. navy \tab aluminum production and processing \tab commercial, industrial, and other intangible assets rental and leasing \tab construction, and mining and oil and gas field machinery manufacturing \tab lumber and other construction materials, merchant wholesalers \tab miscellaneous nonmetallic mineral product manufacturing \tab nonferrous metal (except aluminum) production and processing \tab coating, engraving, heat treating and allied activities \tab fishing, hunting, and trapping \tab internet publishing and broadcasting and web search portals \tab other consumer goods rental \tab railroad rolling stock manufacturing \tab other information services \tab pipeline transportation \tab barber shops \tab other transportation equipment manufacturing \tab car washes \tab footwear and leather goods repair \tab nonmetallic mineral mining and quarrying \tab sewage treatment facilities \tab electronic auctions \tab active duty military personnel, all other \tab electronic shopping \tab miscellaneous petroleum and coal products \tab software publishers \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1296 \tab 8 \tab 2 \tab 2 \tab 8 \tab 3 \tab 1 \tab 1 \tab 9 \tab 2 \tab 6 \tab 6 \tab 1 \tab 1 \tab 3 \tab 4 \tab 2 \tab 1 \tab 3 \tab 25 \tab 26 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab 14 \tab 3 \tab 24 \tab 1 \tab 20 \tab 1 \tab 5 \tab 9 \tab 1 \tab 1 \tab 3 \tab 17 \tab 27 \tab 69 \tab 1 \tab 61 \tab 2 \tab 6 \tab 21 \tab 2 \tab 1 \tab 7 \tab 6 \tab 159 \tab 4 \tab 17 \tab 24 \tab 1 \tab 1 \tab 8 \tab 7 \tab 1 \tab 1 \tab 9 \tab 2 \tab 4 \tab 6 \tab 7 \tab 1 \tab 3 \tab 5 \tab 47 \tab 2 \tab 1 \tab 7 \tab 126 \tab 1 \tab 1 \tab 5 \tab 6 \tab 22 \tab 1 \tab 5 \tab 7 \tab 9 \tab 9 \tab 1 \tab 2 \tab 14 \tab 3 \tab 2 \tab 3 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 8 \tab 7 \tab 2 \tab 1 \tab 1 \tab 2 \tab 22 \tab 1 \tab 1 \tab 1 \tab 17 \tab 1 \tab 3 \tab 7 \tab 59 \tab 7 \tab 3 \tab 1 \tab 15 \tab 9 \tab 2 \tab 1 \tab 41 \tab 1 \tab 7 \tab 6 \tab 16 \tab 1 \tab 4 \tab 1 \tab 10 \tab 6 \tab 2 \tab 2 \tab 2 \tab 4 \tab 5 \tab 2 \tab 1 \tab 1 \tab 14 \tab 2 \tab 1 \tab 2 \tab 10 \tab 5 \tab 11 \tab 1 \tab 13 \tab 90 \tab 1 \tab 1 \tab 1 \tab 1 \tab 15 \tab 1 \tab 13 \tab 5 \tab 4 \tab 91 \tab 5 \tab 3 \tab 11 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 12 \tab 3 \tab 2 \tab 3 \tab 2 \tab 3 \tab 5 \tab 5 \tab 3 \tab 2 \tab 4 \tab 2 \tab 3 \tab 1 \tab 6 \tab 2 \tab 8 \tab 4 \tab 22 \tab 4 \tab 1 \tab 1 \tab 4 \tab 2 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 13 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1124 \tab 11 \tab 5 \tab 1 \tab 11 \tab 1 \tab - \tab 1 \tab 12 \tab 2 \tab 5 \tab 6 \tab 2 \tab 1 \tab 4 \tab 7 \tab - \tab 2 \tab 7 \tab 35 \tab 27 \tab 3 \tab 1 \tab 3 \tab 3 \tab 3 \tab 2 \tab - \tab 17 \tab 8 \tab 11 \tab - \tab 26 \tab 2 \tab 3 \tab 5 \tab - \tab 6 \tab 3 \tab 14 \tab 31 \tab 70 \tab - \tab 39 \tab - \tab 6 \tab 24 \tab 3 \tab 1 \tab 9 \tab 11 \tab 207 \tab 4 \tab 14 \tab 25 \tab 1 \tab - \tab 4 \tab 9 \tab - \tab - \tab 9 \tab 2 \tab 3 \tab 11 \tab 2 \tab 4 \tab 4 \tab 9 \tab 42 \tab 2 \tab 2 \tab 10 \tab 137 \tab 1 \tab - \tab 4 \tab 6 \tab 16 \tab 2 \tab 4 \tab 6 \tab 9 \tab 10 \tab 2 \tab 2 \tab 9 \tab - \tab - \tab - \tab 11 \tab 4 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 11 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 13 \tab 1 \tab - \tab 1 \tab 15 \tab - \tab 2 \tab 4 \tab 49 \tab 1 \tab 4 \tab 1 \tab 8 \tab 9 \tab 2 \tab 1 \tab 52 \tab 1 \tab 7 \tab 4 \tab 21 \tab - \tab - \tab 1 \tab 4 \tab 7 \tab 3 \tab 3 \tab 5 \tab 3 \tab 6 \tab - \tab 5 \tab 3 \tab 11 \tab 7 \tab 4 \tab 4 \tab 12 \tab 2 \tab 13 \tab - \tab 9 \tab 82 \tab - \tab 2 \tab 2 \tab 1 \tab 23 \tab 2 \tab 11 \tab 10 \tab 1 \tab 91 \tab 8 \tab 3 \tab 3 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 19 \tab 3 \tab 3 \tab - \tab 2 \tab 3 \tab 6 \tab 3 \tab 6 \tab 3 \tab 5 \tab 2 \tab 2 \tab 1 \tab 5 \tab 1 \tab 3 \tab 1 \tab 26 \tab 6 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 23 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 8 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1182 \tab 6 \tab 6 \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 6 \tab 1 \tab 3 \tab 10 \tab 5 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 6 \tab 32 \tab 24 \tab 2 \tab - \tab - \tab 5 \tab 6 \tab 1 \tab 3 \tab 29 \tab 5 \tab 13 \tab - \tab 19 \tab - \tab 1 \tab 10 \tab - \tab 3 \tab 1 \tab 16 \tab 21 \tab 66 \tab 3 \tab 47 \tab - \tab 4 \tab 11 \tab - \tab 2 \tab 8 \tab 16 \tab 193 \tab 3 \tab 20 \tab 16 \tab 1 \tab 1 \tab 8 \tab 17 \tab 1 \tab 1 \tab 6 \tab 1 \tab 3 \tab 8 \tab 2 \tab 1 \tab 5 \tab 3 \tab 40 \tab 2 \tab 3 \tab 4 \tab 141 \tab 3 \tab 1 \tab 3 \tab 12 \tab 17 \tab - \tab 4 \tab 1 \tab 12 \tab 7 \tab - \tab 1 \tab 7 \tab 5 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 11 \tab 9 \tab 5 \tab 2 \tab - \tab - \tab 11 \tab 2 \tab 6 \tab - \tab 16 \tab - \tab - \tab 7 \tab 38 \tab 4 \tab 1 \tab - \tab 16 \tab 13 \tab - \tab 1 \tab 50 \tab - \tab 5 \tab 3 \tab 15 \tab 1 \tab 4 \tab - \tab 3 \tab 6 \tab 6 \tab 3 \tab 3 \tab 4 \tab 3 \tab - \tab 2 \tab 7 \tab 12 \tab - \tab 2 \tab 4 \tab 3 \tab 1 \tab 8 \tab - \tab 7 \tab 71 \tab 1 \tab 1 \tab 1 \tab 2 \tab 21 \tab - \tab 5 \tab 5 \tab 6 \tab 92 \tab 6 \tab 5 \tab 2 \tab - \tab 1 \tab 4 \tab 1 \tab 3 \tab 12 \tab 3 \tab 3 \tab 4 \tab - \tab 4 \tab 5 \tab 6 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 4 \tab 3 \tab 14 \tab 3 \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 15 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 2 \tab 3 \tab 5 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 7 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1148 \tab 9 \tab 8 \tab 1 \tab 15 \tab 3 \tab 1 \tab 3 \tab 7 \tab - \tab 5 \tab 6 \tab - \tab 2 \tab 4 \tab 4 \tab - \tab 1 \tab 5 \tab 21 \tab 24 \tab - \tab 2 \tab 6 \tab 4 \tab 6 \tab - \tab - \tab 21 \tab 5 \tab 25 \tab 1 \tab 29 \tab 1 \tab 1 \tab 4 \tab - \tab 4 \tab 1 \tab 10 \tab 26 \tab 61 \tab 4 \tab 34 \tab 1 \tab 8 \tab 11 \tab 1 \tab 1 \tab 7 \tab 11 \tab 176 \tab 4 \tab 14 \tab 18 \tab - \tab 1 \tab 5 \tab 12 \tab - \tab 1 \tab 7 \tab 1 \tab 5 \tab 5 \tab 5 \tab 2 \tab 7 \tab 5 \tab 36 \tab 3 \tab 5 \tab 7 \tab 147 \tab 3 \tab - \tab 3 \tab 18 \tab 20 \tab 3 \tab 2 \tab 6 \tab 11 \tab 2 \tab - \tab - \tab 12 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 9 \tab 5 \tab - \tab 3 \tab - \tab 1 \tab 13 \tab 1 \tab - \tab 3 \tab 14 \tab 3 \tab 1 \tab 9 \tab 43 \tab 3 \tab 8 \tab - \tab 12 \tab 22 \tab 1 \tab 1 \tab 37 \tab - \tab 12 \tab 4 \tab 15 \tab - \tab 2 \tab - \tab 5 \tab 10 \tab 4 \tab 2 \tab 2 \tab 3 \tab 7 \tab 1 \tab 4 \tab 2 \tab 10 \tab - \tab 4 \tab 2 \tab 8 \tab 4 \tab 15 \tab - \tab 11 \tab 85 \tab 1 \tab 2 \tab 1 \tab 2 \tab 24 \tab 1 \tab 8 \tab 7 \tab 2 \tab 98 \tab - \tab 3 \tab 4 \tab - \tab 3 \tab 6 \tab - \tab 1 \tab 11 \tab - \tab 2 \tab 2 \tab 1 \tab 5 \tab 4 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 3 \tab 20 \tab 3 \tab - \tab 1 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 14 \tab 5 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 3 \tab 3 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1104 \tab 5 \tab 5 \tab 1 \tab 6 \tab - \tab - \tab 3 \tab 8 \tab - \tab 5 \tab 8 \tab 2 \tab 1 \tab 7 \tab 4 \tab - \tab 4 \tab 5 \tab 27 \tab 18 \tab - \tab 4 \tab 4 \tab 7 \tab 5 \tab 1 \tab 2 \tab 22 \tab 5 \tab 13 \tab 1 \tab 30 \tab 1 \tab - \tab 10 \tab - \tab 5 \tab 1 \tab 15 \tab 18 \tab 53 \tab 1 \tab 58 \tab 2 \tab 8 \tab 16 \tab 2 \tab - \tab 10 \tab 7 \tab 185 \tab 1 \tab 23 \tab 15 \tab 1 \tab - \tab 3 \tab 7 \tab 2 \tab - \tab 6 \tab - \tab 3 \tab 15 \tab 3 \tab 3 \tab 3 \tab 8 \tab 31 \tab 2 \tab 1 \tab 8 \tab 132 \tab 4 \tab 1 \tab 4 \tab 8 \tab 20 \tab - \tab 1 \tab 2 \tab 15 \tab 2 \tab - \tab 2 \tab 11 \tab 1 \tab - \tab 4 \tab 7 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 5 \tab 8 \tab - \tab - \tab 2 \tab - \tab 17 \tab 1 \tab 1 \tab 1 \tab 12 \tab 1 \tab 2 \tab 11 \tab 40 \tab 1 \tab 1 \tab - \tab 15 \tab 20 \tab 2 \tab 3 \tab 42 \tab 1 \tab 10 \tab 2 \tab 24 \tab - \tab 3 \tab 1 \tab 6 \tab 5 \tab 4 \tab - \tab 1 \tab 4 \tab 5 \tab 1 \tab 2 \tab 2 \tab 6 \tab - \tab 1 \tab 2 \tab - \tab 6 \tab 12 \tab 1 \tab 7 \tab 79 \tab 2 \tab 2 \tab 3 \tab 6 \tab 17 \tab - \tab 6 \tab 3 \tab 4 \tab 110 \tab 3 \tab 7 \tab 4 \tab - \tab 2 \tab 1 \tab 4 \tab 2 \tab 8 \tab 4 \tab 1 \tab 1 \tab 2 \tab 3 \tab 5 \tab 3 \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 4 \tab 20 \tab 5 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 6 \tab - \tab - \tab 21 \tab 3 \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab 3 \tab - \tab 1 \tab 4 \tab 2 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1050 \tab 8 \tab 3 \tab 2 \tab 15 \tab 2 \tab - \tab 4 \tab 4 \tab 1 \tab 4 \tab 6 \tab 5 \tab - \tab 2 \tab 8 \tab 1 \tab 1 \tab 1 \tab 34 \tab 16 \tab 2 \tab 2 \tab 2 \tab 3 \tab 4 \tab - \tab 2 \tab 32 \tab 6 \tab 26 \tab - \tab 24 \tab - \tab 1 \tab 2 \tab - \tab 6 \tab 4 \tab 12 \tab 28 \tab 34 \tab 2 \tab 56 \tab - \tab 6 \tab 12 \tab 6 \tab 1 \tab 4 \tab 6 \tab 230 \tab 2 \tab 16 \tab 10 \tab 1 \tab - \tab 3 \tab 8 \tab - \tab 1 \tab 5 \tab 1 \tab 3 \tab 9 \tab 4 \tab 4 \tab 3 \tab 5 \tab 34 \tab 2 \tab 2 \tab 7 \tab 143 \tab 3 \tab 1 \tab 5 \tab 10 \tab 21 \tab 4 \tab 2 \tab 2 \tab 15 \tab 11 \tab - \tab 3 \tab 22 \tab 7 \tab - \tab 1 \tab 4 \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 3 \tab 8 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 8 \tab - \tab 1 \tab 2 \tab 13 \tab - \tab 5 \tab 8 \tab 79 \tab 2 \tab 6 \tab 1 \tab 15 \tab 13 \tab - \tab - \tab 43 \tab 1 \tab 16 \tab - \tab 20 \tab 1 \tab - \tab 1 \tab 4 \tab 5 \tab 5 \tab 4 \tab 2 \tab 7 \tab 6 \tab - \tab 4 \tab 5 \tab 6 \tab - \tab 1 \tab 1 \tab 10 \tab 4 \tab 8 \tab - \tab 8 \tab 76 \tab - \tab 2 \tab 3 \tab 1 \tab 37 \tab 3 \tab 7 \tab 9 \tab 1 \tab 104 \tab 7 \tab 7 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 10 \tab 2 \tab 2 \tab - \tab 2 \tab 3 \tab 2 \tab 5 \tab 3 \tab 3 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 19 \tab 5 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab - \tab - \tab 15 \tab - \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 4 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 4 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2636 \tab 6 \tab 3 \tab 4 \tab 11 \tab 1 \tab - \tab 4 \tab 5 \tab 1 \tab 9 \tab 9 \tab 2 \tab - \tab 4 \tab 4 \tab 1 \tab 1 \tab 3 \tab 20 \tab 21 \tab 1 \tab 2 \tab 4 \tab 7 \tab 4 \tab - \tab 2 \tab 28 \tab 6 \tab 17 \tab 1 \tab 37 \tab 1 \tab 3 \tab - \tab - \tab 5 \tab 5 \tab 19 \tab 30 \tab 51 \tab 6 \tab 54 \tab - \tab 8 \tab 12 \tab 7 \tab 1 \tab 8 \tab 12 \tab 230 \tab 5 \tab 16 \tab 13 \tab - \tab - \tab 4 \tab 6 \tab - \tab - \tab 8 \tab - \tab 5 \tab 10 \tab 10 \tab 4 \tab 3 \tab 6 \tab 33 \tab 1 \tab 1 \tab 8 \tab 147 \tab 1 \tab 1 \tab 5 \tab 14 \tab 19 \tab 1 \tab 3 \tab 4 \tab 18 \tab 14 \tab 1 \tab - \tab 19 \tab 6 \tab - \tab 1 \tab 6 \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab 12 \tab 10 \tab 2 \tab 1 \tab 1 \tab 3 \tab 16 \tab - \tab 1 \tab 1 \tab 17 \tab 1 \tab 5 \tab 8 \tab 60 \tab 3 \tab 3 \tab - \tab 31 \tab 19 \tab - \tab - \tab 52 \tab 3 \tab 8 \tab 2 \tab 19 \tab - \tab 1 \tab 2 \tab 4 \tab 7 \tab 4 \tab 1 \tab - \tab 7 \tab 5 \tab - \tab 1 \tab 3 \tab 9 \tab 1 \tab 3 \tab 5 \tab 8 \tab 6 \tab 7 \tab - \tab 8 \tab 88 \tab 5 \tab 2 \tab 2 \tab 1 \tab 23 \tab - \tab 8 \tab 9 \tab 6 \tab 110 \tab 7 \tab 2 \tab 11 \tab - \tab 2 \tab 4 \tab 3 \tab 1 \tab 10 \tab 2 \tab 2 \tab - \tab 1 \tab 9 \tab 4 \tab 4 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 4 \tab 5 \tab 32 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 16 \tab 3 \tab - \tab - \tab - \tab 2 \tab 2 \tab 5 \tab - \tab - \tab 2 \tab 5 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 708 \tab 6 \tab 4 \tab 1 \tab 8 \tab 3 \tab 1 \tab 2 \tab 3 \tab - \tab 5 \tab 4 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab 7 \tab 26 \tab 19 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab - \tab - \tab 25 \tab 6 \tab 16 \tab - \tab 14 \tab - \tab 2 \tab 3 \tab - \tab 2 \tab - \tab 22 \tab 23 \tab 34 \tab - \tab 31 \tab - \tab 5 \tab 8 \tab 2 \tab 1 \tab 5 \tab 6 \tab 142 \tab 5 \tab 11 \tab 11 \tab 1 \tab - \tab - \tab 10 \tab - \tab - \tab 12 \tab 2 \tab 3 \tab 7 \tab 9 \tab 1 \tab 12 \tab 3 \tab 23 \tab 2 \tab - \tab 12 \tab 93 \tab 2 \tab 1 \tab 2 \tab 5 \tab 21 \tab 1 \tab 2 \tab 5 \tab 20 \tab 6 \tab - \tab - \tab 10 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab 13 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 6 \tab 32 \tab 1 \tab 1 \tab - \tab 5 \tab 7 \tab - \tab - \tab 50 \tab - \tab 10 \tab 1 \tab 5 \tab - \tab 4 \tab 2 \tab 2 \tab 4 \tab 3 \tab - \tab 2 \tab 1 \tab 8 \tab - \tab 1 \tab 3 \tab 8 \tab 1 \tab - \tab 1 \tab 12 \tab 2 \tab 11 \tab - \tab 6 \tab 58 \tab 2 \tab 4 \tab 1 \tab 1 \tab 26 \tab 2 \tab 3 \tab 5 \tab - \tab 86 \tab 5 \tab 1 \tab 6 \tab - \tab 3 \tab 3 \tab 3 \tab 2 \tab 4 \tab 4 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 4 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 5 \tab 20 \tab 5 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 15 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 640 \tab 6 \tab 3 \tab 1 \tab 10 \tab 4 \tab 1 \tab 1 \tab 5 \tab 1 \tab 7 \tab 6 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 3 \tab 27 \tab 21 \tab 1 \tab 2 \tab 3 \tab 3 \tab 3 \tab - \tab 1 \tab 17 \tab 7 \tab 10 \tab 1 \tab 32 \tab - \tab 1 \tab 3 \tab - \tab 6 \tab 1 \tab 10 \tab 19 \tab 39 \tab 3 \tab 42 \tab - \tab 3 \tab 8 \tab 1 \tab 1 \tab 7 \tab 5 \tab 161 \tab 5 \tab 10 \tab 6 \tab 1 \tab 1 \tab - \tab 6 \tab - \tab - \tab 8 \tab 3 \tab 2 \tab 10 \tab 7 \tab 4 \tab 6 \tab 7 \tab 36 \tab 2 \tab 1 \tab 16 \tab 105 \tab 1 \tab 1 \tab 2 \tab 7 \tab 23 \tab 2 \tab 2 \tab 2 \tab 17 \tab 5 \tab 1 \tab 2 \tab 5 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 4 \tab 19 \tab - \tab 2 \tab 3 \tab 11 \tab 1 \tab 1 \tab 4 \tab 41 \tab 3 \tab 2 \tab - \tab 1 \tab 9 \tab - \tab 1 \tab 40 \tab - \tab 13 \tab - \tab 18 \tab 1 \tab 2 \tab 2 \tab 6 \tab 7 \tab 4 \tab 3 \tab - \tab 3 \tab 6 \tab - \tab 2 \tab 3 \tab 7 \tab - \tab - \tab 3 \tab 6 \tab 4 \tab 10 \tab - \tab 6 \tab 46 \tab 1 \tab - \tab 1 \tab 2 \tab 22 \tab 2 \tab 10 \tab 4 \tab 2 \tab 108 \tab 2 \tab 4 \tab 3 \tab - \tab - \tab 6 \tab 2 \tab 2 \tab 5 \tab 2 \tab - \tab - \tab 4 \tab 3 \tab 3 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab - \tab 24 \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 19 \tab 6 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 6 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 606 \tab 10 \tab 4 \tab 1 \tab 9 \tab - \tab 1 \tab 1 \tab 5 \tab - \tab 1 \tab 17 \tab - \tab 2 \tab 5 \tab 4 \tab 1 \tab 1 \tab 5 \tab 29 \tab 24 \tab 2 \tab 2 \tab 3 \tab 2 \tab 6 \tab 1 \tab - \tab 30 \tab 9 \tab 13 \tab - \tab 19 \tab 1 \tab - \tab 5 \tab 4 \tab 1 \tab 2 \tab 9 \tab 14 \tab 34 \tab 2 \tab 36 \tab - \tab 6 \tab 15 \tab 1 \tab - \tab 5 \tab 3 \tab 130 \tab 6 \tab 13 \tab 9 \tab 1 \tab - \tab 3 \tab 6 \tab - \tab - \tab 8 \tab 1 \tab 6 \tab 1 \tab 5 \tab 1 \tab 2 \tab 4 \tab 35 \tab 2 \tab 5 \tab 12 \tab 95 \tab 1 \tab 1 \tab 5 \tab 11 \tab 25 \tab 1 \tab 2 \tab 2 \tab 18 \tab 2 \tab 2 \tab 3 \tab 13 \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 3 \tab 1 \tab 2 \tab 14 \tab - \tab 4 \tab 2 \tab 2 \tab 2 \tab 3 \tab 4 \tab 27 \tab 2 \tab 5 \tab - \tab 15 \tab 11 \tab 1 \tab - \tab 41 \tab - \tab 4 \tab 3 \tab 16 \tab - \tab - \tab 3 \tab 5 \tab 1 \tab 4 \tab - \tab 4 \tab 3 \tab 9 \tab - \tab 1 \tab 6 \tab 6 \tab 1 \tab 1 \tab 4 \tab 3 \tab 4 \tab 5 \tab 1 \tab 4 \tab 53 \tab - \tab 1 \tab 2 \tab 1 \tab 15 \tab - \tab 8 \tab 8 \tab 3 \tab 72 \tab 3 \tab 4 \tab 2 \tab - \tab 2 \tab 4 \tab 3 \tab 3 \tab 19 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 2 \tab 13 \tab 4 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 19 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 24 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 845 \tab 9 \tab 8 \tab - \tab 13 \tab 1 \tab 1 \tab 5 \tab 3 \tab 4 \tab 5 \tab 11 \tab 1 \tab - \tab 5 \tab 4 \tab 1 \tab 4 \tab 4 \tab 39 \tab 27 \tab - \tab 1 \tab 3 \tab 4 \tab 8 \tab 1 \tab - \tab 30 \tab 3 \tab 7 \tab - \tab 31 \tab 1 \tab 2 \tab 5 \tab 1 \tab 5 \tab 1 \tab 15 \tab 20 \tab 40 \tab 3 \tab 46 \tab - \tab 9 \tab 9 \tab 5 \tab 2 \tab 6 \tab 6 \tab 195 \tab 1 \tab 12 \tab 9 \tab 2 \tab - \tab 2 \tab 4 \tab - \tab - \tab 7 \tab - \tab 3 \tab 3 \tab 5 \tab 2 \tab 4 \tab 4 \tab 42 \tab 2 \tab 3 \tab 15 \tab 132 \tab 2 \tab 3 \tab 5 \tab 18 \tab 32 \tab 2 \tab 2 \tab 2 \tab 19 \tab 2 \tab 1 \tab 1 \tab 12 \tab 6 \tab - \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab 11 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 10 \tab 1 \tab 2 \tab 1 \tab 12 \tab 1 \tab 1 \tab 7 \tab 29 \tab 4 \tab 1 \tab 2 \tab 7 \tab 8 \tab 1 \tab 2 \tab 61 \tab 2 \tab 14 \tab 3 \tab 23 \tab - \tab 2 \tab 1 \tab 7 \tab 7 \tab 5 \tab - \tab 2 \tab 5 \tab 12 \tab - \tab 2 \tab 1 \tab 7 \tab 1 \tab 2 \tab 1 \tab 9 \tab 1 \tab 5 \tab 1 \tab 6 \tab 66 \tab 1 \tab 4 \tab - \tab - \tab 28 \tab 1 \tab 7 \tab 7 \tab 3 \tab 101 \tab 6 \tab 1 \tab 12 \tab 1 \tab 5 \tab 4 \tab 4 \tab 3 \tab 11 \tab 3 \tab - \tab 1 \tab 3 \tab 4 \tab 7 \tab 5 \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 6 \tab 1 \tab 3 \tab 4 \tab 25 \tab 4 \tab 1 \tab - \tab - \tab 4 \tab - \tab 2 \tab 7 \tab 3 \tab - \tab 14 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 8 \tab 1 \tab - \tab - \tab - \tab 1 \tab 5 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 820 \tab 8 \tab 5 \tab 3 \tab 13 \tab 1 \tab - \tab 1 \tab 10 \tab 1 \tab 5 \tab 19 \tab 1 \tab 3 \tab 2 \tab 4 \tab 1 \tab 2 \tab 6 \tab 44 \tab 25 \tab - \tab 2 \tab 3 \tab 6 \tab 6 \tab 3 \tab 1 \tab 33 \tab 10 \tab 20 \tab - \tab 29 \tab 1 \tab 2 \tab 3 \tab 3 \tab 2 \tab 3 \tab 17 \tab 28 \tab 30 \tab 3 \tab 30 \tab - \tab 7 \tab 16 \tab 7 \tab 1 \tab 4 \tab 9 \tab 245 \tab 6 \tab 21 \tab 5 \tab - \tab 1 \tab 1 \tab 7 \tab - \tab 3 \tab 5 \tab - \tab 6 \tab 4 \tab 10 \tab 3 \tab 3 \tab 12 \tab 42 \tab - \tab 2 \tab 19 \tab 174 \tab 3 \tab 1 \tab 4 \tab 19 \tab 34 \tab 2 \tab 2 \tab - \tab 17 \tab 6 \tab - \tab 4 \tab 20 \tab 9 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 11 \tab 7 \tab - \tab - \tab 4 \tab 3 \tab 19 \tab 1 \tab 4 \tab 1 \tab 19 \tab 1 \tab 2 \tab 9 \tab 71 \tab 6 \tab - \tab 1 \tab 14 \tab 21 \tab - \tab 3 \tab 64 \tab 1 \tab 12 \tab 3 \tab 34 \tab - \tab 1 \tab 2 \tab 10 \tab 3 \tab 8 \tab 1 \tab - \tab 8 \tab 11 \tab - \tab - \tab 6 \tab 9 \tab - \tab - \tab 4 \tab 7 \tab 2 \tab 9 \tab - \tab 10 \tab 83 \tab 1 \tab 1 \tab 2 \tab - \tab 25 \tab - \tab 10 \tab 18 \tab 1 \tab 113 \tab 6 \tab 8 \tab 4 \tab - \tab 6 \tab 3 \tab 8 \tab 5 \tab 21 \tab 2 \tab - \tab - \tab 3 \tab 4 \tab 6 \tab 3 \tab - \tab 3 \tab 6 \tab 1 \tab 2 \tab 1 \tab 5 \tab 2 \tab 2 \tab 7 \tab 40 \tab 7 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 4 \tab 2 \tab - \tab 14 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 3 \tab 1 \tab 1 \tab 4 \tab 7 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 7 \tab 3 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 671 \tab 18 \tab 2 \tab 3 \tab 13 \tab 2 \tab - \tab 5 \tab 6 \tab 1 \tab 4 \tab 14 \tab 2 \tab 1 \tab 5 \tab 3 \tab 1 \tab 3 \tab 6 \tab 43 \tab 22 \tab - \tab 4 \tab 7 \tab 10 \tab 5 \tab 1 \tab 1 \tab 33 \tab 5 \tab 25 \tab - \tab 17 \tab - \tab 1 \tab 2 \tab 1 \tab 4 \tab 6 \tab 16 \tab 15 \tab 25 \tab 1 \tab 25 \tab - \tab 3 \tab 14 \tab 4 \tab - \tab 6 \tab 6 \tab 211 \tab 6 \tab 9 \tab 6 \tab 1 \tab - \tab - \tab 9 \tab 1 \tab - \tab 8 \tab 1 \tab 5 \tab 8 \tab 10 \tab 3 \tab 3 \tab 9 \tab 35 \tab 3 \tab 3 \tab 16 \tab 126 \tab 5 \tab - \tab 3 \tab 17 \tab 23 \tab 1 \tab - \tab 4 \tab 14 \tab 4 \tab 2 \tab 2 \tab 12 \tab 3 \tab - \tab - \tab 5 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 7 \tab 4 \tab - \tab 4 \tab - \tab 4 \tab 11 \tab 2 \tab 3 \tab 4 \tab 13 \tab 2 \tab 3 \tab 4 \tab 27 \tab 8 \tab 6 \tab - \tab 9 \tab 39 \tab - \tab - \tab 55 \tab 2 \tab 12 \tab 3 \tab 29 \tab 1 \tab - \tab 2 \tab 12 \tab 5 \tab 11 \tab 4 \tab 2 \tab 5 \tab 9 \tab 1 \tab - \tab 4 \tab 10 \tab 1 \tab 1 \tab 5 \tab 5 \tab 2 \tab 15 \tab - \tab 5 \tab 38 \tab - \tab 2 \tab 1 \tab 1 \tab 21 \tab - \tab 11 \tab 5 \tab 2 \tab 89 \tab 3 \tab 8 \tab 4 \tab - \tab 3 \tab 3 \tab 1 \tab 5 \tab 24 \tab 2 \tab 1 \tab 4 \tab 1 \tab 6 \tab 3 \tab 3 \tab 1 \tab - \tab 5 \tab - \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab 5 \tab 20 \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab 12 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1178 \tab 18 \tab 9 \tab 1 \tab 26 \tab 2 \tab - \tab 5 \tab 4 \tab - \tab 11 \tab 11 \tab 4 \tab 2 \tab 8 \tab 8 \tab - \tab 5 \tab 5 \tab 72 \tab 48 \tab 4 \tab 4 \tab 5 \tab 7 \tab 14 \tab 2 \tab 4 \tab 50 \tab 13 \tab 35 \tab 1 \tab 56 \tab - \tab 2 \tab 6 \tab 2 \tab 5 \tab 11 \tab 20 \tab 16 \tab 59 \tab 5 \tab 43 \tab - \tab 11 \tab 20 \tab 7 \tab - \tab 8 \tab 11 \tab 347 \tab 6 \tab 40 \tab 5 \tab 4 \tab - \tab 1 \tab 5 \tab 1 \tab - \tab 13 \tab 2 \tab 17 \tab 10 \tab 4 \tab 7 \tab 3 \tab 13 \tab 51 \tab 4 \tab 7 \tab 23 \tab 213 \tab 2 \tab 3 \tab 9 \tab 36 \tab 43 \tab - \tab 2 \tab 6 \tab 23 \tab 3 \tab 1 \tab - \tab 40 \tab 14 \tab - \tab - \tab 4 \tab 3 \tab 2 \tab - \tab - \tab - \tab 4 \tab 16 \tab 17 \tab 3 \tab 4 \tab - \tab - \tab 25 \tab 3 \tab 4 \tab 5 \tab 19 \tab 3 \tab 1 \tab 10 \tab 109 \tab 10 \tab - \tab 1 \tab 22 \tab 40 \tab 1 \tab 2 \tab 78 \tab 2 \tab 24 \tab 3 \tab 45 \tab - \tab 2 \tab 4 \tab 8 \tab 12 \tab 24 \tab 7 \tab 10 \tab 14 \tab 17 \tab 1 \tab 5 \tab 7 \tab 9 \tab 1 \tab - \tab 6 \tab 12 \tab 11 \tab 15 \tab - \tab 11 \tab 38 \tab 1 \tab 5 \tab - \tab 2 \tab 32 \tab 1 \tab 15 \tab 8 \tab - \tab 170 \tab 3 \tab 4 \tab 4 \tab 1 \tab 5 \tab 7 \tab 6 \tab 8 \tab 47 \tab 2 \tab 2 \tab 3 \tab 1 \tab 5 \tab 7 \tab 5 \tab - \tab 4 \tab 8 \tab 4 \tab 2 \tab 1 \tab 5 \tab 1 \tab - \tab 6 \tab 29 \tab 7 \tab - \tab - \tab - \tab 7 \tab - \tab - \tab 7 \tab 1 \tab - \tab 22 \tab 5 \tab 1 \tab - \tab 5 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 6 \tab - \tab 4 \tab 1 \tab 4 \tab 9 \tab - \tab 4 \tab 3 \tab - \tab 1 \tab 6 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 11 \tab 1 \tab 5 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 951 \tab 16 \tab 3 \tab 4 \tab 18 \tab 4 \tab 1 \tab 6 \tab 7 \tab 1 \tab 8 \tab 16 \tab 5 \tab 2 \tab 12 \tab 11 \tab 1 \tab 6 \tab 4 \tab 55 \tab 31 \tab 3 \tab 1 \tab 5 \tab 7 \tab 15 \tab 5 \tab - \tab 44 \tab 9 \tab 38 \tab - \tab 41 \tab - \tab 1 \tab 4 \tab 1 \tab 2 \tab 10 \tab 25 \tab 34 \tab 43 \tab 1 \tab 43 \tab - \tab 11 \tab 27 \tab 4 \tab - \tab 3 \tab 14 \tab 297 \tab 5 \tab 26 \tab 5 \tab - \tab - \tab 2 \tab 5 \tab - \tab - \tab 6 \tab 1 \tab 11 \tab 3 \tab 8 \tab 4 \tab 4 \tab 6 \tab 46 \tab 1 \tab 6 \tab 28 \tab 202 \tab 5 \tab 1 \tab 7 \tab 38 \tab 33 \tab 2 \tab 5 \tab 7 \tab 19 \tab 2 \tab 2 \tab 2 \tab 35 \tab 6 \tab - \tab 2 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 13 \tab 9 \tab 4 \tab 4 \tab - \tab 3 \tab 23 \tab - \tab 5 \tab 6 \tab 20 \tab 2 \tab 3 \tab 10 \tab 108 \tab 15 \tab 1 \tab 1 \tab 20 \tab 32 \tab - \tab - \tab 52 \tab 4 \tab 17 \tab 11 \tab 38 \tab - \tab 3 \tab 5 \tab 13 \tab 5 \tab 29 \tab 3 \tab 6 \tab 16 \tab 15 \tab 1 \tab 1 \tab 2 \tab 8 \tab - \tab 1 \tab 3 \tab 17 \tab 4 \tab 14 \tab - \tab 9 \tab 47 \tab - \tab 3 \tab 1 \tab 1 \tab 31 \tab 1 \tab 10 \tab 15 \tab 5 \tab 147 \tab 5 \tab 4 \tab 4 \tab - \tab 8 \tab 4 \tab 10 \tab 5 \tab 51 \tab 5 \tab 5 \tab 4 \tab 5 \tab 4 \tab 6 \tab 2 \tab 1 \tab 3 \tab 8 \tab 1 \tab - \tab 1 \tab 5 \tab 1 \tab 5 \tab 3 \tab 40 \tab 7 \tab - \tab - \tab - \tab 6 \tab 1 \tab 2 \tab 7 \tab 1 \tab - \tab 32 \tab 7 \tab - \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 7 \tab - \tab - \tab - \tab 3 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 5 \tab 5 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab 3 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 3544 \cr
   Total \tab 45347 \tab 144 \tab 70 \tab 26 \tab 186 \tab 29 \tab 10 \tab 46 \tab 94 \tab 15 \tab 83 \tab 149 \tab 32 \tab 17 \tab 67 \tab 70 \tab 10 \tab 35 \tab 70 \tab 529 \tab 373 \tab 22 \tab 30 \tab 50 \tab 71 \tab 90 \tab 21 \tab 17 \tab 425 \tab 100 \tab 293 \tab 6 \tab 424 \tab 9 \tab 25 \tab 71 \tab 13 \tab 57 \tab 52 \tab 237 \tab 350 \tab 708 \tab 35 \tab 645 \tab 5 \tab 101 \tab 224 \tab 52 \tab 12 \tab 97 \tab 129 \tab 3108 \tab 63 \tab 262 \tab 177 \tab 15 \tab 5 \tab 44 \tab 118 \tab 6 \tab 7 \tab 117 \tab 17 \tab 79 \tab 110 \tab 91 \tab 44 \tab 65 \tab 99 \tab 573 \tab 30 \tab 42 \tab 192 \tab 2113 \tab 37 \tab 16 \tab 66 \tab 225 \tab 369 \tab 22 \tab 38 \tab 56 \tab 236 \tab 85 \tab 13 \tab 24 \tab 241 \tab 70 \tab 5 \tab 15 \tab 63 \tab 19 \tab 16 \tab 5 \tab 15 \tab 10 \tab 35 \tab 128 \tab 98 \tab 21 \tab 25 \tab 12 \tab 28 \tab 234 \tab 15 \tab 37 \tab 32 \tab 203 \tab 19 \tab 33 \tab 108 \tab 812 \tab 70 \tab 42 \tab 8 \tab 205 \tab 272 \tab 10 \tab 15 \tab 758 \tab 18 \tab 171 \tab 48 \tab 338 \tab 5 \tab 28 \tab 27 \tab 99 \tab 90 \tab 116 \tab 33 \tab 41 \tab 87 \tab 124 \tab 7 \tab 31 \tab 55 \tab 132 \tab 15 \tab 21 \tab 47 \tab 122 \tab 58 \tab 158 \tab 4 \tab 120 \tab 1000 \tab 16 \tab 32 \tab 21 \tab 22 \tab 360 \tab 14 \tab 132 \tab 118 \tab 40 \tab 1582 \tab 69 \tab 64 \tab 75 \tab 4 \tab 44 \tab 56 \tab 48 \tab 44 \tab 264 \tab 37 \tab 25 \tab 24 \tab 32 \tab 59 \tab 69 \tab 51 \tab 33 \tab 27 \tab 56 \tab 17 \tab 15 \tab 16 \tab 57 \tab 17 \tab 43 \tab 53 \tab 364 \tab 70 \tab 8 \tab 8 \tab 8 \tab 41 \tab 6 \tab 15 \tab 56 \tab 12 \tab 2 \tab 264 \tab 49 \tab 5 \tab 4 \tab 23 \tab 13 \tab 21 \tab 19 \tab 7 \tab 16 \tab 7 \tab 48 \tab 9 \tab 17 \tab 11 \tab 24 \tab 54 \tab 6 \tab 18 \tab 11 \tab 16 \tab 25 \tab 67 \tab 37 \tab 8 \tab 17 \tab 3 \tab 13 \tab 12 \tab 14 \tab 32 \tab 16 \tab 10 \tab 1 \tab 4 \tab 16 \tab 4 \tab 10 \tab 5 \tab 4 \tab 3 \tab 20 \tab 7 \tab 18 \tab 14 \tab 5 \tab 16 \tab 2 \tab 6 \tab 8 \tab 7 \tab 26 \tab 4 \tab 5 \tab 5 \tab 3 \tab 8 \tab 2 \tab 6 \tab 3 \tab 4 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 6 \tab 2 \tab 4 \tab 1 \tab 3 \tab 2 \tab 1 \tab 9 \tab 1 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{170} crop production
\item \code{180} animal production
\item \code{190} forestry, except logging
\item \code{270} logging
\item \code{280} fishing, hunting, and trapping
\item \code{290} support activities for agriculture and forestry
\item \code{370} oil and gas extraction
\item \code{380} coal mining
\item \code{390} metal ore mining
\item \code{470} nonmetallic mineral mining and quarrying
\item \code{480} not specified type of mining
\item \code{490} support activities for mining
\item \code{570} electric power generation, transmission, and distribution
\item \code{580} natural gas distribution
\item \code{590} electric and gas, and other combinations
\item \code{670} water, steam, air-conditioning, and irrigation systems
\item \code{680} sewage treatment facilities
\item \code{690} not specified utilities
\item \code{770} construction
\item \code{1070} animal food, grain, and oilseed milling
\item \code{1080} sugar and confectionery products
\item \code{1090} fruit and vegetable preserving and specialty food manufacturing
\item \code{1170} dairy product manufacturing
\item \code{1180} animal slaughtering and processing
\item \code{1190} retail bakeries
\item \code{1270} bakeries, except retail
\item \code{1280} seafood and other miscellaneous foods, n.e.c.
\item \code{1290} not specified food industries
\item \code{1370} beverage manufacturing
\item \code{1390} tobacco manufacturing
\item \code{1470} fiber, yarn, and thread mills
\item \code{1480} fabric mills, except knitting mills
\item \code{1490} textile and fabric finishing and coating mills
\item \code{1570} carpet and rug mills
\item \code{1590} textile product mills, except carpet and rug
\item \code{1670} knitting fabric mills, and apparel knitting mills
\item \code{1680} cut and sew apparel manufacturing
\item \code{1690} apparel accessories and other apparel manufacturing
\item \code{1770} footwear manufacturing
\item \code{1790} leather tanning and finishing and other allied products manufacturing
\item \code{1870} pulp, paper, and paperboard mills
\item \code{1880} paperboard containers and boxes
\item \code{1890} miscellaneous paper and pulp products
\item \code{1990} printing and related support activities
\item \code{2070} petroleum refining
\item \code{2090} miscellaneous petroleum and coal products
\item \code{2170} resin, synthetic rubber and fibers, and filaments manufacturing
\item \code{2180} agricultural chemical manufacturing
\item \code{2190} pharmaceutical and medicine manufacturing
\item \code{2270} paint, coating, and adhesive manufacturing
\item \code{2280} soap, cleaning compound, and cosmetics manufacturing
\item \code{2290} industrial and miscellaneous chemicals
\item \code{2370} plastics product manufacturing
\item \code{2380} tire manufacturing
\item \code{2390} rubber product, except tire, manufacturing
\item \code{2470} pottery, ceramics, and plumbing fixture manufacturing
\item \code{2480} structural clay product manufacturing
\item \code{2490} glass and glass product manufacturing
\item \code{2570} cement, concrete, lime, and gypsum product manufacturing
\item \code{2590} miscellaneous nonmetallic mineral product manufacturing
\item \code{2670} iron and steel mills and steel product manufacturing
\item \code{2680} aluminum production and processing
\item \code{2690} nonferrous metal (except aluminum) production and processing
\item \code{2770} foundries
\item \code{2780} metal forgings and stampings
\item \code{2790} cutlery and hand tool manufacturing
\item \code{2870} structural metals, and boiler, tank, and shipping container manufacturing
\item \code{2880} machine shops; turned product; screw, nut, and bolt manufacturing
\item \code{2890} coating, engraving, heat treating and allied activities
\item \code{2970} ordnance
\item \code{2980} miscellaneous fabricated metal products manufacturing
\item \code{2990} not specified metal industries
\item \code{3070} agricultural implement manufacturing
\item \code{3080} construction, and mining and oil and gas field machinery manufacturing
\item \code{3090} commercial and service industry machinery manufacturing
\item \code{3170} metalworking machinery manufacturing
\item \code{3180} engines, turbines, and power transmission equipment manufacturing
\item \code{3190} machinery manufacturing, n.e.c.
\item \code{3290} not specified machinery manufacturing
\item \code{3360} computer and peripheral equipment manufacturing
\item \code{3370} communications, and audio and video equipment manufacturing
\item \code{3380} navigational, measuring, electromedical, and control instruments manufacturing
\item \code{3390} electronic component and product manufacturing, n.e.c.
\item \code{3470} household appliance manufacturing
\item \code{3490} electrical lighting and electrical equipment manufacturing, and other electrical component manufacturing, n.e.c.
\item \code{3570} motor vehicles and motor vehicle equipment manufacturing
\item \code{3580} aircraft and parts manufacturing
\item \code{3590} aerospace product and parts manufacturing
\item \code{3670} railroad rolling stock manufacturing
\item \code{3680} ship and boat building
\item \code{3690} other transportation equipment manufacturing
\item \code{3770} sawmills and wood preservation
\item \code{3780} veneer, plywood, and engineered wood products
\item \code{3790} prefabricated wood buildings and mobile homes
\item \code{3870} miscellaneous wood products
\item \code{3890} furniture and related product manufacturing
\item \code{3960} medical equipment and supplies manufacturing
\item \code{3970} sporting and athletic goods, and doll, toy and game manufacturing
\item \code{3980} miscellaneous manufacturing, n.e.c.
\item \code{3990} not specified manufacturing industries
\item \code{4070} motor vehicles, parts and supplies, merchant wholesalers
\item \code{4080} furniture and home furnishing, merchant wholesalers
\item \code{4090} lumber and other construction materials, merchant wholesalers
\item \code{4170} professional and commercial equipment and supplies, merchant wholesalers
\item \code{4180} metals and minerals, except petroleum, merchant wholesalers
\item \code{4190} electrical and electronic goods, merchant wholesalers
\item \code{4260} hardware, plumbing and heating equipment, and supplies, merchant wholesalers
\item \code{4270} machinery, equipment, and supplies, merchant wholesalers
\item \code{4280} recyclable material, merchant wholesalers
\item \code{4290} miscellaneous durable goods, merchant wholesalers
\item \code{4370} paper and paper products, merchant wholesalers
\item \code{4380} drugs, sundries, and chemical and allied products, merchant wholesalers
\item \code{4390} apparel, fabrics, and notions, merchant wholesalers
\item \code{4470} groceries and related products, merchant wholesalers
\item \code{4480} farm product raw materials, merchant wholesalers
\item \code{4490} petroleum and petroleum products, merchant wholesalers
\item \code{4560} alcoholic beverages, merchant wholesalers
\item \code{4570} farm supplies, merchant wholesalers
\item \code{4580} miscellaneous nondurable goods, merchant wholesalers
\item \code{4585} wholesale electronic markets, agents and brokers
\item \code{4590} not specified wholesale trade
\item \code{4670} automobile dealers
\item \code{4680} other motor vehicle dealers
\item \code{4690} auto parts, accessories, and tire stores
\item \code{4770} furniture and home furnishings stores
\item \code{4780} household appliance stores
\item \code{4790} radio, tv, and computer stores
\item \code{4870} building material and supplies dealers
\item \code{4880} hardware stores
\item \code{4890} lawn and garden equipment and supplies stores
\item \code{4970} grocery stores
\item \code{4980} specialty food stores
\item \code{4990} beer, wine, and liquor stores
\item \code{5070} pharmacies and drug stores
\item \code{5080} health and personal care, except drug, stores
\item \code{5090} gasoline stations
\item \code{5170} clothing stores
\item \code{5180} shoe stores
\item \code{5190} jewelry, luggage, and leather goods stores
\item \code{5270} sporting goods, camera, and hobby and toy stores
\item \code{5280} sewing, needlework, and piece goods stores
\item \code{5290} music stores
\item \code{5370} book stores and news dealers
\item \code{5380} department stores and discount stores
\item \code{5390} miscellaneous general merchandise stores
\item \code{5470} retail florists
\item \code{5480} office supplies and stationery stores
\item \code{5490} used merchandise stores
\item \code{5570} gift, novelty, and souvenir shops
\item \code{5580} miscellaneous retail stores
\item \code{5590} electronic shopping
\item \code{5591} electronic auctions
\item \code{5592} mail order houses
\item \code{5670} vending machine operators
\item \code{5680} fuel dealers
\item \code{5690} other direct selling establishments
\item \code{5790} not specified retail trade
\item \code{6070} air transportation
\item \code{6080} rail transportation
\item \code{6090} water transportation
\item \code{6170} truck transportation
\item \code{6180} bus service and urban transit
\item \code{6190} taxi and limousine service
\item \code{6270} pipeline transportation
\item \code{6280} scenic and sightseeing transportation
\item \code{6290} services incidental to transportation
\item \code{6370} postal service
\item \code{6380} couriers and messengers
\item \code{6390} warehousing and storage
\item \code{6470} newspaper publishers
\item \code{6480} periodical, book, and directory publishers
\item \code{6490} software publishers
\item \code{6570} motion pictures and video industries
\item \code{6590} sound recording industries
\item \code{6670} radio and television broadcasting and cable subscription programming
\item \code{6672} internet publishing and broadcasting and web search portals
\item \code{6680} wired telecommunications carriers
\item \code{6690} other telecommunications services
\item \code{6695} data processing, hosting, and related services
\item \code{6770} libraries and archives
\item \code{6780} other information services
\item \code{6870} banking and related activities
\item \code{6880} savings institutions, including credit unions
\item \code{6890} non-depository credit and related activities
\item \code{6970} securities, commodities, funds, trusts, and other financial investments
\item \code{6990} insurance carriers and related activities
\item \code{7070} real estate
\item \code{7080} automotive equipment rental and leasing
\item \code{7170} video tape and disk rental
\item \code{7180} other consumer goods rental
\item \code{7190} commercial, industrial, and other intangible assets rental and leasing
\item \code{7270} legal services
\item \code{7280} accounting, tax preparation, bookkeeping, and payroll services
\item \code{7290} architectural, engineering, and related services
\item \code{7370} specialized design services
\item \code{7380} computer systems design and related services
\item \code{7390} management, scientific, and technical consulting services
\item \code{7460} scientific research and development services
\item \code{7470} advertising and related services
\item \code{7480} veterinary services
\item \code{7490} other professional, scientific, and technical services
\item \code{7570} management of companies and enterprises
\item \code{7580} employment services
\item \code{7590} business support services
\item \code{7670} travel arrangements and reservation services
\item \code{7680} investigation and security services
\item \code{7690} services to buildings and dwellings
\item \code{7770} landscaping services
\item \code{7780} other administrative and other support services
\item \code{7790} waste management and remediation services
\item \code{7860} elementary and secondary schools
\item \code{7870} colleges and universities, including junior colleges
\item \code{7880} business, technical, and trade schools and training
\item \code{7890} other schools and instruction, and educational support services
\item \code{7970} offices of physicians
\item \code{7980} offices of dentists
\item \code{7990} offices of chiropractors
\item \code{8070} offices of optometrists
\item \code{8080} offices of other health practitioners
\item \code{8090} outpatient care centers
\item \code{8170} home health care services
\item \code{8180} other health care services
\item \code{8190} hospitals
\item \code{8270} nursing care facilities
\item \code{8290} residential care facilities, without nursing
\item \code{8370} individual and family services
\item \code{8380} community food and housing, and emergency services
\item \code{8390} vocational rehabilitation services
\item \code{8470} child day care services
\item \code{8560} independent artists, performing arts, spectator sports, and related industries
\item \code{8570} museums, art galleries, historical sites, and similar institutions
\item \code{8580} bowling centers
\item \code{8590} other amusement, gambling, and recreation industries
\item \code{8660} traveler accommodation
\item \code{8670} recreational vehicle parks and camps, and rooming and boarding houses
\item \code{8680} restaurants and other food services
\item \code{8690} drinking places, alcoholic beverages
\item \code{8770} automotive repair and maintenance
\item \code{8780} car washes
\item \code{8790} electronic and precision equipment repair and maintenance
\item \code{8870} commercial and industrial machinery and equipment repair and maintenance
\item \code{8880} personal and household goods repair and maintenance
\item \code{8890} footwear and leather goods repair
\item \code{8970} barber shops
\item \code{8980} beauty salons
\item \code{8990} nail salons and other personal care services
\item \code{9070} drycleaning and laundry services
\item \code{9080} funeral homes, and cemeteries and crematories
\item \code{9090} other personal services
\item \code{9160} religious organizations
\item \code{9170} civic, social, advocacy organizations, and grantmaking and giving services
\item \code{9180} labor unions
\item \code{9190} business, professional, political, and similar organizations
\item \code{9290} private households
\item \code{9370} executive offices and legislative bodies
\item \code{9380} public finance activities
\item \code{9390} other general government and support
\item \code{9470} justice, public order, and safety activities
\item \code{9480} administration of human resource programs
\item \code{9490} administration of environmental quality and housing programs
\item \code{9570} administration of economic programs and space research
\item \code{9590} national security and international affairs
\item \code{9670} u.s. army
\item \code{9680} u.s. air force
\item \code{9690} u.s. navy
\item \code{9770} u.s. marine corps
\item \code{9780} u.s. coast guard
\item \code{9790} u.s. armed forces/n.s.
\item \code{9870} active duty military personnel, all other
\item \code{9890} uncodeable
\item \code{9997} uncodeable
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
