% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_09.R
\name{maindstatus}
\alias{maindstatus}
\title{Maind10 coding status}
\source{
General Social Survey https://gss.norc.org
}
\description{
maindstatus
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab found, iap \tab not found, iap \tab found, coded \tab found, no answer \tab not found, coded \tab not found, no answer \tab not found, no code assigned \tab iap \tab Total \cr
   1972 \tab 702 \tab 911 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1234 \tab 270 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1217 \tab 267 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 976 \tab 514 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1230 \tab 269 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 817 \tab 713 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1518 \tab 14 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1293 \tab 175 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1696 \tab 164 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1475 \tab 124 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1458 \tab 15 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1296 \tab 238 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1425 \tab 45 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1511 \tab 308 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1460 \tab 21 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1420 \tab 117 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1358 \tab 14 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1499 \tab 18 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1545 \tab 61 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1266 \tab 30 \tab 1614 \tab 58 \tab 17 \tab 3 \tab 4 \tab - \tab 2992 \cr
   1996 \tab 1115 \tab 9 \tab 1720 \tab 48 \tab 8 \tab 1 \tab 3 \tab - \tab 2904 \cr
   1998 \tab 1162 \tab 20 \tab 1602 \tab 37 \tab 9 \tab 1 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 1148 \tab - \tab 1625 \tab 44 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1104 \tab - \tab 1619 \tab 42 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1050 \tab - \tab 1683 \tab 79 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2636 \tab - \tab 1814 \tab 60 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 708 \tab - \tab 1283 \tab 32 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 640 \tab - \tab 1363 \tab 41 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 35959 \tab 4317 \tab 14323 \tab 441 \tab 34 \tab 5 \tab 8 \tab 17303 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} found, iap
\item \code{2} found, no answer
\item \code{3} found, coded
\item \code{4} not found, iap
\item \code{5} not found, no answer
\item \code{6} not found, coded
\item \code{7} not found, no code assigned
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
