% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_595.R
\name{maisc681}
\alias{maisc681}
\title{R's mothers occupation, 1980 census}
\source{
General Social Survey https://gss.norc.org
}
\description{
maisc681
}
\details{
Question 1640. Occupation of respondent's mother (International Standard Classification Codes-1968)
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 110 \tab 1100 \tab 1219 \tab 1310 \tab 1320 \tab 1330 \tab 1340 \tab 1350 \tab 1390 \tab 1393 \tab 140 \tab 1410 \tab 1490 \tab 1590 \tab 1610 \tab 1621 \tab 1639 \tab 1730 \tab 1910 \tab 1930 \tab 1939 \tab 1940 \tab 2010 \tab 2110 \tab 2190 \tab 2192 \tab 3000 \tab 3100 \tab 320 \tab 3211 \tab 3219 \tab 3220 \tab 3311 \tab 3313 \tab 3319 \tab 3399 \tab 3420 \tab 3520 \tab 3590 \tab 3700 \tab 3709 \tab 3809 \tab 3919 \tab 3920 \tab 3939 \tab 3940 \tab 3949 \tab 3951 \tab 3959 \tab 3990 \tab 4210 \tab 4220 \tab 4320 \tab 4400 \tab 4410 \tab 4411 \tab 4412 \tab 4419 \tab 4431 \tab 4432 \tab 4519 \tab 4520 \tab 4521 \tab 4523 \tab 510 \tab 5200 \tab 530 \tab 5300 \tab 5312 \tab 5319 \tab 5321 \tab 5322 \tab 5329 \tab 540 \tab 5400 \tab 5401 \tab 549 \tab 5500 \tab 5510 \tab 5520 \tab 5600 \tab 5700 \tab 5702 \tab 5810 \tab 5829 \tab 5890 \tab 5900 \tab 5920 \tab 5990 \tab 5991 \tab 5999 \tab 6001 \tab 610 \tab 6110 \tab 620 \tab 6210 \tab 6270 \tab 630 \tab 6319 \tab 640 \tab 650 \tab 670 \tab 690 \tab 7000 \tab 710 \tab 7240 \tab 7430 \tab 7520 \tab 7540 \tab 770 \tab 7730 \tab 7760 \tab 7790 \tab 7910 \tab 7919 \tab 7950 \tab 7960 \tab 7990 \tab 8020 \tab 830 \tab 8340 \tab 8420 \tab 849 \tab 8493 \tab 8499 \tab 8530 \tab 8559 \tab 8720 \tab 9210 \tab 9220 \tab 9250 \tab 9260 \tab 9270 \tab 9390 \tab 9430 \tab 9490 \tab 9499 \tab 9594 \tab 9712 \tab 9714 \tab 9851 \tab 9852 \tab 9859 \tab 9891 \tab 9979 \tab 9990 \tab 9996 \tab 1519 \tab 1593 \tab 1710 \tab 1790 \tab 1921 \tab 1999 \tab 230 \tab 240 \tab 3102 \tab 3391 \tab 390 \tab 3911 \tab 3999 \tab 4513 \tab 5320 \tab 5709 \tab 6000 \tab 6240 \tab 7410 \tab 760 \tab 761 \tab 790 \tab 8010 \tab 810 \tab 8120 \tab 8351 \tab 8421 \tab 8422 \tab 8494 \tab 8520 \tab 8620 \tab 8800 \tab 9310 \tab 9599 \tab 9690 \tab 9730 \tab 9790 \tab 1799 \tab 2020 \tab 2030 \tab 210 \tab 340 \tab 3410 \tab 4420 \tab 5895 \tab 5992 \tab 7119 \tab 7320 \tab 769 \tab 8110 \tab 8190 \tab 8440 \tab 8910 \tab 8950 \tab 1229 \tab 1924 \tab 220 \tab 250 \tab 8490 \tab 8560 \tab 3801 \tab 4222 \tab 5891 \tab 5998 \tab 6320 \tab 759 \tab 8430 \tab 8610 \tab 900 \tab 120 \tab 8329 \tab 8330 \tab 8730 \tab 9510 \tab 9530 \tab 9595 \tab 9710 \tab 280 \tab 6410 \tab 1720 \tab 310 \tab 3800 \tab 3930 \tab 5821 \tab 7280 \tab 131 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1371 \tab 1 \tab 12 \tab 6 \tab 5 \tab 13 \tab 63 \tab 4 \tab 5 \tab 5 \tab 16 \tab 1 \tab 1 \tab 2 \tab 4 \tab 9 \tab 6 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 7 \tab 2 \tab 3 \tab 29 \tab 3 \tab 16 \tab 1 \tab 1 \tab 121 \tab 6 \tab 4 \tab 57 \tab 8 \tab 49 \tab 1 \tab 3 \tab 3 \tab 5 \tab 6 \tab 1 \tab 12 \tab 3 \tab 3 \tab 22 \tab 2 \tab 10 \tab 4 \tab 3 \tab 12 \tab 35 \tab 2 \tab 6 \tab 2 \tab 10 \tab 9 \tab 1 \tab 3 \tab 4 \tab 4 \tab 64 \tab 1 \tab 13 \tab 2 \tab 2 \tab 7 \tab 1 \tab 12 \tab 1 \tab 63 \tab 6 \tab 2 \tab 55 \tab 4 \tab 19 \tab 25 \tab 1 \tab 24 \tab 1 \tab 75 \tab 26 \tab 2 \tab 25 \tab 1 \tab 2 \tab 4 \tab 3 \tab 1 \tab 1 \tab 2 \tab 65 \tab 1 \tab 2 \tab 11 \tab 4 \tab 18 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 5 \tab 10 \tab 84 \tab 1 \tab 1 \tab 13 \tab 10 \tab 2 \tab 4 \tab 1 \tab 2 \tab 18 \tab 1 \tab 65 \tab 1 \tab 3 \tab 5 \tab 4 \tab 19 \tab 1 \tab 1 \tab 32 \tab 5 \tab 15 \tab 2 \tab 4 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 6 \tab 4 \tab 16 \tab 1 \tab 9 \tab 29 \tab 7 \tab 1 \tab 2 \tab 2 \tab 2 \tab 33 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1176 \tab - \tab 16 \tab 1 \tab 5 \tab 24 \tab 71 \tab 8 \tab 3 \tab 10 \tab 6 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 6 \tab 1 \tab 1 \tab 4 \tab 1 \tab 12 \tab 9 \tab 5 \tab 7 \tab 80 \tab 1 \tab 6 \tab - \tab 1 \tab 141 \tab 8 \tab 6 \tab 54 \tab 13 \tab 54 \tab 2 \tab 1 \tab 1 \tab 6 \tab 8 \tab 6 \tab 15 \tab 1 \tab 6 \tab 15 \tab 1 \tab 13 \tab 3 \tab 1 \tab 7 \tab 26 \tab 1 \tab 4 \tab 3 \tab 7 \tab 11 \tab - \tab 3 \tab 3 \tab 3 \tab 61 \tab 1 \tab 9 \tab - \tab - \tab 32 \tab - \tab 14 \tab 4 \tab 50 \tab 3 \tab 2 \tab 48 \tab 5 \tab 25 \tab 20 \tab 3 \tab 20 \tab 3 \tab 64 \tab 33 \tab - \tab 26 \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 62 \tab - \tab 3 \tab 12 \tab 3 \tab 16 \tab - \tab 1 \tab - \tab 6 \tab - \tab 1 \tab 6 \tab 13 \tab 102 \tab 1 \tab - \tab 3 \tab 8 \tab 1 \tab 1 \tab 1 \tab 2 \tab 24 \tab 4 \tab 59 \tab - \tab 9 \tab 4 \tab 1 \tab 20 \tab - \tab 1 \tab 40 \tab 6 \tab 5 \tab - \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 2 \tab 12 \tab 1 \tab 9 \tab 20 \tab 7 \tab 2 \tab 1 \tab 2 \tab 5 \tab 41 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 5 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1222 \tab 2 \tab 13 \tab 4 \tab 3 \tab 28 \tab 75 \tab 12 \tab 5 \tab 9 \tab 12 \tab - \tab - \tab - \tab 4 \tab 2 \tab 6 \tab - \tab - \tab 6 \tab - \tab 16 \tab 2 \tab 4 \tab 3 \tab 63 \tab - \tab 6 \tab - \tab 1 \tab 119 \tab 4 \tab 7 \tab 39 \tab 13 \tab 38 \tab 6 \tab 2 \tab 1 \tab 4 \tab 6 \tab 3 \tab 13 \tab 6 \tab 6 \tab 25 \tab - \tab 18 \tab 2 \tab - \tab 12 \tab 37 \tab 3 \tab 3 \tab 1 \tab 9 \tab 11 \tab - \tab - \tab 4 \tab 3 \tab 72 \tab - \tab 2 \tab - \tab 1 \tab 14 \tab - \tab 12 \tab 3 \tab 42 \tab 2 \tab 2 \tab 43 \tab 1 \tab 15 \tab 28 \tab 5 \tab 28 \tab 2 \tab 45 \tab 15 \tab - \tab 23 \tab - \tab - \tab 4 \tab 5 \tab 1 \tab 1 \tab - \tab 57 \tab - \tab 4 \tab 11 \tab 1 \tab 11 \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 5 \tab 7 \tab 107 \tab 1 \tab 1 \tab 9 \tab 2 \tab 3 \tab 2 \tab 5 \tab - \tab 38 \tab 2 \tab 47 \tab 1 \tab 3 \tab 7 \tab - \tab 30 \tab - \tab 2 \tab 46 \tab 3 \tab 6 \tab 2 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 3 \tab 3 \tab 9 \tab 1 \tab 2 \tab 25 \tab 8 \tab 2 \tab 1 \tab 2 \tab - \tab 39 \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 9 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1194 \tab 1 \tab 11 \tab 2 \tab 9 \tab 16 \tab 74 \tab 11 \tab 3 \tab 2 \tab 10 \tab 1 \tab - \tab 3 \tab 3 \tab 3 \tab 3 \tab - \tab 1 \tab 8 \tab - \tab 13 \tab 6 \tab 4 \tab 6 \tab 80 \tab 3 \tab 2 \tab - \tab 1 \tab 131 \tab 7 \tab 4 \tab 32 \tab 10 \tab 47 \tab 5 \tab - \tab - \tab 1 \tab 12 \tab 3 \tab 11 \tab 2 \tab 3 \tab 28 \tab - \tab 7 \tab 3 \tab 1 \tab 14 \tab 29 \tab - \tab 5 \tab 2 \tab 5 \tab 14 \tab - \tab 5 \tab 6 \tab 1 \tab 75 \tab 1 \tab 9 \tab 2 \tab - \tab 7 \tab - \tab 23 \tab 1 \tab 54 \tab 6 \tab 4 \tab 41 \tab 1 \tab 29 \tab 26 \tab 6 \tab 17 \tab 2 \tab 66 \tab 15 \tab 2 \tab 26 \tab - \tab 2 \tab 7 \tab 4 \tab 1 \tab 2 \tab 1 \tab 58 \tab - \tab 2 \tab 11 \tab 3 \tab 18 \tab 1 \tab 3 \tab - \tab 8 \tab - \tab 4 \tab 4 \tab 6 \tab 95 \tab - \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab 5 \tab 3 \tab 28 \tab 1 \tab 54 \tab - \tab 3 \tab 8 \tab 4 \tab 43 \tab - \tab 2 \tab 35 \tab 2 \tab 8 \tab 1 \tab 3 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 7 \tab 1 \tab 4 \tab 20 \tab 9 \tab 1 \tab 2 \tab - \tab - \tab 15 \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1146 \tab 1 \tab 9 \tab 5 \tab 7 \tab 13 \tab 66 \tab 10 \tab 5 \tab 10 \tab 10 \tab 1 \tab 1 \tab - \tab 4 \tab 4 \tab 12 \tab 1 \tab - \tab 8 \tab - \tab 12 \tab 5 \tab 4 \tab 4 \tab 76 \tab 2 \tab 4 \tab - \tab 1 \tab 150 \tab 5 \tab 1 \tab 33 \tab 7 \tab 33 \tab 4 \tab 1 \tab 2 \tab 3 \tab 7 \tab 2 \tab 13 \tab 2 \tab 6 \tab 35 \tab - \tab 11 \tab 8 \tab 1 \tab 12 \tab 39 \tab 2 \tab 4 \tab 1 \tab 6 \tab 9 \tab 3 \tab 6 \tab 5 \tab 3 \tab 84 \tab - \tab 3 \tab 3 \tab 1 \tab 10 \tab - \tab 14 \tab 4 \tab 47 \tab 5 \tab 9 \tab 63 \tab 2 \tab 21 \tab 26 \tab 1 \tab 26 \tab 1 \tab 64 \tab 12 \tab - \tab 19 \tab - \tab 5 \tab 6 \tab 2 \tab - \tab 1 \tab 1 \tab 56 \tab - \tab 4 \tab 6 \tab 3 \tab 10 \tab 2 \tab - \tab - \tab 7 \tab - \tab - \tab 1 \tab 13 \tab 95 \tab 2 \tab - \tab 5 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 16 \tab 1 \tab 42 \tab 2 \tab 8 \tab 2 \tab 3 \tab 44 \tab - \tab 4 \tab 43 \tab 1 \tab 14 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 9 \tab - \tab 4 \tab 16 \tab 10 \tab 2 \tab 1 \tab 2 \tab 2 \tab 11 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1116 \tab 1 \tab 25 \tab 9 \tab 8 \tab 25 \tab 98 \tab 12 \tab 4 \tab 10 \tab 15 \tab 1 \tab - \tab 2 \tab 6 \tab - \tab 7 \tab 1 \tab 1 \tab 9 \tab 2 \tab 9 \tab 4 \tab 10 \tab 3 \tab 76 \tab - \tab 6 \tab 1 \tab - \tab 111 \tab 5 \tab 5 \tab 25 \tab 17 \tab 34 \tab 5 \tab 2 \tab 1 \tab 7 \tab 5 \tab 1 \tab 9 \tab 5 \tab 4 \tab 31 \tab - \tab 14 \tab 1 \tab 4 \tab 16 \tab 33 \tab 5 \tab 4 \tab 2 \tab 6 \tab 20 \tab 1 \tab 9 \tab 5 \tab 2 \tab 81 \tab - \tab 6 \tab - \tab - \tab 14 \tab 2 \tab 21 \tab 3 \tab 49 \tab 2 \tab 5 \tab 62 \tab 8 \tab 30 \tab 31 \tab 4 \tab 27 \tab 1 \tab 49 \tab 15 \tab 1 \tab 16 \tab - \tab 2 \tab 6 \tab 5 \tab - \tab - \tab 2 \tab 54 \tab - \tab 4 \tab 6 \tab 8 \tab 23 \tab 1 \tab 3 \tab - \tab 9 \tab - \tab 2 \tab 5 \tab 1 \tab 98 \tab - \tab 1 \tab 3 \tab 4 \tab 3 \tab 3 \tab 6 \tab 2 \tab 11 \tab 4 \tab 44 \tab - \tab 5 \tab 2 \tab 3 \tab 33 \tab - \tab 4 \tab 37 \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 5 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 4 \tab 6 \tab 1 \tab 5 \tab 11 \tab 9 \tab 2 \tab 1 \tab 2 \tab 2 \tab 24 \tab - \tab - \tab 1 \tab 6 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2715 \tab 1 \tab 22 \tab 6 \tab 10 \tab 54 \tab 65 \tab 9 \tab 1 \tab 9 \tab 7 \tab 1 \tab - \tab - \tab 6 \tab 6 \tab 7 \tab - \tab - \tab 8 \tab 1 \tab 17 \tab 9 \tab 3 \tab 5 \tab 76 \tab 3 \tab 12 \tab 1 \tab - \tab 128 \tab 3 \tab 6 \tab 44 \tab 9 \tab 33 \tab 5 \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab 10 \tab 6 \tab 11 \tab 23 \tab 2 \tab 10 \tab 3 \tab 2 \tab 14 \tab 26 \tab 3 \tab 5 \tab - \tab 7 \tab 7 \tab - \tab 5 \tab 13 \tab 2 \tab 77 \tab - \tab 7 \tab 1 \tab 2 \tab 12 \tab - \tab 18 \tab 6 \tab 58 \tab 5 \tab 3 \tab 48 \tab 11 \tab 25 \tab 42 \tab 3 \tab 28 \tab 2 \tab 75 \tab 18 \tab - \tab 20 \tab - \tab 5 \tab 2 \tab 5 \tab - \tab 1 \tab 1 \tab 41 \tab 1 \tab 2 \tab 12 \tab 14 \tab 19 \tab 1 \tab 2 \tab - \tab 6 \tab - \tab 2 \tab 4 \tab 8 \tab 115 \tab 2 \tab - \tab 2 \tab 5 \tab - \tab 3 \tab 4 \tab 2 \tab 25 \tab 2 \tab 40 \tab - \tab 6 \tab 2 \tab 4 \tab 38 \tab 1 \tab 7 \tab 42 \tab 6 \tab 5 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 8 \tab 2 \tab 7 \tab 19 \tab 14 \tab 1 \tab 1 \tab - \tab - \tab 42 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 5 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 8 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 738 \tab - \tab 14 \tab 6 \tab 4 \tab 25 \tab 49 \tab 9 \tab 1 \tab 1 \tab 8 \tab 2 \tab - \tab - \tab 6 \tab - \tab 2 \tab - \tab - \tab 4 \tab 4 \tab 7 \tab 7 \tab 5 \tab 3 \tab 62 \tab - \tab 5 \tab - \tab - \tab 101 \tab 3 \tab 6 \tab 36 \tab 13 \tab 33 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 6 \tab 9 \tab - \tab 6 \tab 25 \tab 1 \tab 5 \tab 3 \tab - \tab 9 \tab 15 \tab 5 \tab 2 \tab 7 \tab 3 \tab 9 \tab - \tab 7 \tab 2 \tab - \tab 51 \tab 1 \tab 4 \tab - \tab 1 \tab 22 \tab - \tab 15 \tab 5 \tab 45 \tab 3 \tab - \tab 37 \tab 3 \tab 24 \tab 25 \tab 1 \tab 18 \tab - \tab 38 \tab 12 \tab 1 \tab 21 \tab - \tab 3 \tab 3 \tab 16 \tab 1 \tab 1 \tab 1 \tab 35 \tab - \tab 2 \tab 13 \tab 6 \tab 13 \tab - \tab 1 \tab 1 \tab 6 \tab - \tab 1 \tab 1 \tab 4 \tab 73 \tab 1 \tab - \tab 3 \tab 4 \tab - \tab 1 \tab 3 \tab 2 \tab 15 \tab 3 \tab 25 \tab - \tab 9 \tab 5 \tab - \tab 32 \tab - \tab 3 \tab 19 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 4 \tab 2 \tab 5 \tab 17 \tab 7 \tab 2 \tab - \tab 1 \tab 1 \tab 13 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 5 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 2023 \cr
   2010 \tab 701 \tab - \tab 20 \tab 3 \tab 6 \tab 29 \tab 63 \tab 4 \tab 1 \tab 10 \tab 4 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 7 \tab - \tab - \tab 9 \tab - \tab 11 \tab 9 \tab 2 \tab 4 \tab 74 \tab 3 \tab 5 \tab - \tab 1 \tab 82 \tab 7 \tab 4 \tab 29 \tab 11 \tab 31 \tab 4 \tab 2 \tab 1 \tab 12 \tab 7 \tab - \tab 11 \tab 2 \tab 9 \tab 20 \tab 2 \tab 7 \tab - \tab 2 \tab 15 \tab 19 \tab 1 \tab 4 \tab 1 \tab 4 \tab 14 \tab - \tab 5 \tab 4 \tab 1 \tab 57 \tab - \tab 5 \tab - \tab 2 \tab 6 \tab - \tab 19 \tab 2 \tab 36 \tab 4 \tab 5 \tab 54 \tab 4 \tab 19 \tab 15 \tab 4 \tab 23 \tab 3 \tab 34 \tab 11 \tab 1 \tab 23 \tab - \tab - \tab 4 \tab 12 \tab - \tab 1 \tab 3 \tab 33 \tab - \tab 7 \tab 9 \tab 8 \tab 13 \tab 3 \tab 1 \tab - \tab 7 \tab 1 \tab 3 \tab 4 \tab 5 \tab 83 \tab 3 \tab - \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 16 \tab 2 \tab 35 \tab 1 \tab 6 \tab 2 \tab 3 \tab 42 \tab - \tab 4 \tab 27 \tab 4 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 5 \tab - \tab 3 \tab 15 \tab 4 \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 5 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 58070 \tab 7 \tab 142 \tab 42 \tab 57 \tab 227 \tab 624 \tab 79 \tab 28 \tab 66 \tab 88 \tab 7 \tab 2 \tab 9 \tab 37 \tab 28 \tab 56 \tab 4 \tab 4 \tab 57 \tab 10 \tab 100 \tab 58 \tab 39 \tab 38 \tab 616 \tab 15 \tab 62 \tab 3 \tab 6 \tab 1084 \tab 48 \tab 43 \tab 349 \tab 101 \tab 352 \tab 33 \tab 13 \tab 12 \tab 41 \tab 62 \tab 23 \tab 103 \tab 27 \tab 54 \tab 224 \tab 8 \tab 95 \tab 27 \tab 14 \tab 111 \tab 259 \tab 22 \tab 37 \tab 19 \tab 57 \tab 104 \tab 5 \tab 43 \tab 46 \tab 19 \tab 622 \tab 4 \tab 58 \tab 8 \tab 9 \tab 124 \tab 3 \tab 148 \tab 29 \tab 444 \tab 36 \tab 32 \tab 451 \tab 39 \tab 207 \tab 238 \tab 28 \tab 211 \tab 15 \tab 510 \tab 157 \tab 7 \tab 199 \tab 1 \tab 21 \tab 38 \tab 54 \tab 5 \tab 8 \tab 13 \tab 461 \tab 2 \tab 30 \tab 91 \tab 50 \tab 141 \tab 13 \tab 13 \tab 2 \tab 55 \tab 3 \tab 14 \tab 35 \tab 67 \tab 852 \tab 11 \tab 4 \tab 42 \tab 40 \tab 11 \tab 19 \tab 29 \tab 16 \tab 191 \tab 20 \tab 411 \tab 5 \tab 52 \tab 37 \tab 22 \tab 301 \tab 2 \tab 28 \tab 321 \tab 35 \tab 55 \tab 5 \tab 23 \tab 7 \tab 17 \tab 1 \tab 10 \tab 10 \tab 5 \tab 19 \tab 28 \tab 76 \tab 9 \tab 48 \tab 172 \tab 75 \tab 14 \tab 9 \tab 11 \tab 12 \tab 222 \tab 1 \tab 1 \tab 7 \tab 16 \tab 5 \tab 9 \tab 14 \tab 7 \tab 2 \tab 7 \tab 13 \tab 3 \tab 31 \tab 10 \tab 5 \tab 10 \tab 4 \tab 3 \tab 6 \tab 8 \tab 3 \tab 5 \tab 17 \tab 5 \tab 1 \tab 3 \tab 3 \tab 3 \tab 5 \tab 9 \tab 1 \tab 1 \tab 3 \tab 3 \tab 3 \tab 8 \tab 1 \tab 7 \tab 2 \tab 1 \tab 1 \tab 8 \tab 9 \tab 5 \tab 4 \tab 20 \tab 4 \tab 2 \tab 1 \tab 13 \tab 5 \tab 3 \tab 1 \tab 5 \tab 3 \tab 5 \tab 3 \tab 4 \tab 3 \tab 3 \tab 4 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 4 \tab 5 \tab 7 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
