% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_596.R
\name{maisco88}
\alias{maisco88}
\title{maisco88}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's mothers occupation, 1980 census & 1988 isco code
}
\details{
Question 1644. Respondent's Mother's Occupation, 1988 Census
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 110 \tab 1110 \tab 1210 \tab 1221 \tab 1223 \tab 1225 \tab 1226 \tab 1227 \tab 1229 \tab 1231 \tab 1232 \tab 1233 \tab 1311 \tab 1315 \tab 1319 \tab 2110 \tab 2113 \tab 2131 \tab 2132 \tab 2148 \tab 2149 \tab 2211 \tab 2221 \tab 2223 \tab 2224 \tab 2229 \tab 2230 \tab 2310 \tab 2320 \tab 2331 \tab 2332 \tab 2340 \tab 2359 \tab 2411 \tab 2419 \tab 2421 \tab 2431 \tab 2432 \tab 2446 \tab 2451 \tab 2452 \tab 2460 \tab 3111 \tab 3114 \tab 3115 \tab 3121 \tab 3122 \tab 3133 \tab 3151 \tab 3152 \tab 3211 \tab 3221 \tab 3223 \tab 3225 \tab 3226 \tab 3229 \tab 3231 \tab 3300 \tab 3410 \tab 3411 \tab 3412 \tab 3413 \tab 3415 \tab 3416 \tab 3417 \tab 3429 \tab 3432 \tab 3442 \tab 3450 \tab 3460 \tab 3471 \tab 3480 \tab 4100 \tab 4110 \tab 4111 \tab 4113 \tab 4114 \tab 4115 \tab 4121 \tab 4122 \tab 4131 \tab 4133 \tab 4141 \tab 4142 \tab 4190 \tab 4211 \tab 4212 \tab 4221 \tab 4222 \tab 4223 \tab 5120 \tab 5122 \tab 5123 \tab 5131 \tab 5132 \tab 5133 \tab 5141 \tab 5143 \tab 5162 \tab 5169 \tab 5200 \tab 5220 \tab 6113 \tab 6141 \tab 7137 \tab 7200 \tab 7212 \tab 7214 \tab 7215 \tab 7222 \tab 7223 \tab 7230 \tab 7243 \tab 7300 \tab 7311 \tab 7313 \tab 7324 \tab 7341 \tab 7411 \tab 7412 \tab 7415 \tab 7433 \tab 7435 \tab 7437 \tab 7442 \tab 7510 \tab 8152 \tab 8200 \tab 8224 \tab 8230 \tab 8231 \tab 8232 \tab 8251 \tab 8253 \tab 8261 \tab 8262 \tab 8263 \tab 8264 \tab 8266 \tab 8269 \tab 8270 \tab 8282 \tab 8290 \tab 8310 \tab 8322 \tab 8323 \tab 9110 \tab 9131 \tab 9132 \tab 9140 \tab 9141 \tab 9142 \tab 9152 \tab 9211 \tab 9312 \tab 9321 \tab 9322 \tab 9333 \tab uncodeable \tab 1220 \tab 1239 \tab 2121 \tab 2143 \tab 2145 \tab 2412 \tab 2444 \tab 2445 \tab 2453 \tab 2455 \tab 3110 \tab 3118 \tab 3132 \tab 3142 \tab 3475 \tab 4143 \tab 4213 \tab 4215 \tab 5139 \tab 6129 \tab 7141 \tab 7231 \tab 7422 \tab 8212 \tab 8333 \tab 8334 \tab 9113 \tab 9151 \tab 9320 \tab 2141 \tab 3224 \tab 3470 \tab 3472 \tab 4132 \tab 5100 \tab 5111 \tab 5163 \tab 7124 \tab 7129 \tab 7232 \tab 7241 \tab 8141 \tab 8162 \tab 8221 \tab 8275 \tab 1234 \tab 1235 \tab 1313 \tab 2142 \tab 2146 \tab 2222 \tab 2422 \tab 3112 \tab 3340 \tab 7245 \tab 8151 \tab 8223 \tab 9153 \tab 2441 \tab 2442 \tab 3131 \tab 3422 \tab 9212 \tab 9311 \tab 1236 \tab 2112 \tab 2213 \tab 2440 \tab 3144 \tab 3212 \tab 3227 \tab 3443 \tab 8100 \tab 8122 \tab 8332 \tab 9213 \tab 7122 \tab 9313 \tab 6124 \tab 1237 \tab 2122 \tab 5161 \tab 3434 \tab 7136 \tab 7213 \tab 8163 \tab 2147 \tab 7132 \tab 9161 \tab 5112 \tab 5113 \tab 7343 \tab 7420 \tab 8340 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1347 \tab 1 \tab 1 \tab 2 \tab 11 \tab 1 \tab 21 \tab 3 \tab 12 \tab 3 \tab 2 \tab 6 \tab 38 \tab 1 \tab 4 \tab 15 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 52 \tab 4 \tab 18 \tab 62 \tab 6 \tab 4 \tab 4 \tab 15 \tab 1 \tab 2 \tab 1 \tab 5 \tab 7 \tab 4 \tab 9 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 19 \tab 5 \tab 8 \tab 5 \tab 3 \tab 3 \tab 1 \tab 96 \tab 15 \tab 2 \tab 1 \tab 5 \tab 9 \tab 9 \tab 4 \tab 2 \tab 3 \tab 13 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 2 \tab 18 \tab 7 \tab 3 \tab 3 \tab 138 \tab 53 \tab 1 \tab 12 \tab 3 \tab 3 \tab 6 \tab 3 \tab 39 \tab 7 \tab 2 \tab 10 \tab 12 \tab 5 \tab 57 \tab 67 \tab 21 \tab 1 \tab 2 \tab 28 \tab 1 \tab 1 \tab 7 \tab 1 \tab 78 \tab 1 \tab 1 \tab 1 \tab 34 \tab 3 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 21 \tab 2 \tab 1 \tab 1 \tab 12 \tab 1 \tab 38 \tab 2 \tab 2 \tab 2 \tab 3 \tab 5 \tab 2 \tab 15 \tab 7 \tab 66 \tab 21 \tab 3 \tab 5 \tab 8 \tab 16 \tab 13 \tab 1 \tab 3 \tab 7 \tab 7 \tab 106 \tab 1 \tab 19 \tab 2 \tab 1 \tab 2 \tab 19 \tab 1 \tab 1 \tab 6 \tab 9 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1157 \tab - \tab - \tab 7 \tab 13 \tab - \tab 32 \tab 1 \tab 18 \tab 6 \tab 5 \tab 1 \tab 33 \tab - \tab 3 \tab 9 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 59 \tab 8 \tab 26 \tab 70 \tab 6 \tab 4 \tab 6 \tab 19 \tab 2 \tab - \tab - \tab 5 \tab 17 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 15 \tab 6 \tab 7 \tab 6 \tab 5 \tab 4 \tab 6 \tab 101 \tab 9 \tab 1 \tab - \tab 3 \tab 14 \tab 10 \tab 1 \tab 3 \tab 5 \tab 10 \tab 1 \tab - \tab - \tab 7 \tab - \tab 3 \tab 15 \tab 6 \tab 8 \tab 5 \tab 151 \tab 58 \tab 2 \tab 16 \tab 3 \tab 7 \tab 12 \tab 5 \tab 40 \tab 12 \tab - \tab 12 \tab 15 \tab 14 \tab 53 \tab 57 \tab 20 \tab 1 \tab 3 \tab 27 \tab 1 \tab 2 \tab 4 \tab 1 \tab 72 \tab - \tab - \tab - \tab 31 \tab 5 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab 4 \tab 3 \tab 17 \tab 3 \tab - \tab 1 \tab 12 \tab 1 \tab 45 \tab 1 \tab 1 \tab - \tab 3 \tab 6 \tab - \tab 4 \tab 9 \tab 71 \tab 33 \tab 5 \tab 9 \tab 5 \tab 8 \tab 15 \tab - \tab 3 \tab 7 \tab 7 \tab 105 \tab 1 \tab 16 \tab 5 \tab 1 \tab - \tab 18 \tab - \tab 4 \tab 6 \tab 10 \tab 1 \tab 2 \tab 6 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1210 \tab 1 \tab - \tab 4 \tab 10 \tab - \tab 25 \tab 2 \tab 12 \tab 11 \tab 7 \tab - \tab 34 \tab - \tab - \tab 13 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab 64 \tab 5 \tab 26 \tab 74 \tab 12 \tab 6 \tab 7 \tab 14 \tab 2 \tab 1 \tab - \tab 7 \tab 17 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 4 \tab - \tab 10 \tab 4 \tab 3 \tab 7 \tab 3 \tab 4 \tab 5 \tab 91 \tab 14 \tab 3 \tab - \tab 1 \tab 10 \tab 12 \tab 3 \tab 1 \tab 3 \tab 12 \tab - \tab 1 \tab - \tab 5 \tab - \tab 1 \tab 23 \tab 14 \tab 5 \tab 3 \tab 125 \tab 43 \tab 2 \tab 19 \tab 2 \tab 2 \tab 7 \tab 6 \tab 42 \tab 11 \tab 1 \tab 14 \tab 17 \tab 5 \tab 47 \tab 54 \tab 27 \tab 1 \tab 4 \tab 24 \tab - \tab - \tab 4 \tab 1 \tab 63 \tab - \tab - \tab 1 \tab 32 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 8 \tab 5 \tab 2 \tab 19 \tab - \tab - \tab 1 \tab 8 \tab - \tab 49 \tab 3 \tab 4 \tab - \tab 2 \tab 5 \tab - \tab 13 \tab 4 \tab 67 \tab 13 \tab 10 \tab 7 \tab 3 \tab 11 \tab 13 \tab - \tab 2 \tab 8 \tab 4 \tab 87 \tab - \tab 13 \tab 2 \tab - \tab - \tab 14 \tab - \tab 3 \tab 16 \tab 9 \tab 3 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1184 \tab - \tab - \tab 3 \tab 11 \tab - \tab 27 \tab - \tab 12 \tab 6 \tab 6 \tab 2 \tab 42 \tab - \tab 1 \tab 16 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 57 \tab 9 \tab 19 \tab 63 \tab 10 \tab 4 \tab 11 \tab 11 \tab 5 \tab 1 \tab - \tab 7 \tab 17 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 8 \tab 4 \tab 3 \tab 8 \tab 10 \tab 7 \tab 4 \tab 90 \tab 12 \tab 2 \tab - \tab 6 \tab 14 \tab 11 \tab 4 \tab - \tab 3 \tab 4 \tab - \tab 2 \tab - \tab 4 \tab 2 \tab - \tab 18 \tab 8 \tab 4 \tab 3 \tab 151 \tab 46 \tab 1 \tab 6 \tab - \tab 5 \tab 15 \tab 12 \tab 37 \tab 10 \tab 1 \tab 4 \tab 11 \tab 5 \tab 56 \tab 56 \tab 28 \tab 1 \tab 3 \tab 26 \tab - \tab - \tab 6 \tab 1 \tab 68 \tab 2 \tab 1 \tab - \tab 30 \tab 5 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 3 \tab 3 \tab 2 \tab 29 \tab 1 \tab - \tab 1 \tab 9 \tab - \tab 42 \tab - \tab - \tab - \tab 1 \tab 5 \tab 2 \tab 5 \tab 3 \tab 56 \tab 15 \tab 5 \tab 6 \tab 8 \tab 7 \tab 11 \tab - \tab 5 \tab 11 \tab 11 \tab 102 \tab 5 \tab 13 \tab 3 \tab 2 \tab 1 \tab 16 \tab - \tab 2 \tab 15 \tab 3 \tab - \tab - \tab 6 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1139 \tab 3 \tab - \tab 3 \tab 4 \tab 1 \tab 25 \tab 2 \tab 18 \tab 9 \tab 5 \tab 3 \tab 38 \tab 1 \tab 2 \tab 13 \tab - \tab 1 \tab 7 \tab 1 \tab - \tab 1 \tab 2 \tab 4 \tab 1 \tab - \tab - \tab 40 \tab 6 \tab 11 \tab 64 \tab 11 \tab 4 \tab 11 \tab 11 \tab 2 \tab 2 \tab 1 \tab 10 \tab 17 \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 11 \tab 3 \tab 1 \tab 1 \tab 7 \tab 3 \tab 5 \tab 104 \tab 16 \tab - \tab 3 \tab 5 \tab 9 \tab 13 \tab 2 \tab 1 \tab 3 \tab 8 \tab - \tab 1 \tab - \tab 12 \tab - \tab 1 \tab 32 \tab 9 \tab 1 \tab 3 \tab 170 \tab 35 \tab 2 \tab 15 \tab 1 \tab 7 \tab 10 \tab 10 \tab 29 \tab 8 \tab - \tab 11 \tab 10 \tab 9 \tab 53 \tab 77 \tab 23 \tab - \tab 6 \tab 20 \tab - \tab 4 \tab 6 \tab - \tab 79 \tab - \tab - \tab - \tab 23 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 3 \tab 2 \tab 13 \tab - \tab 2 \tab 1 \tab 13 \tab - \tab 49 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 4 \tab 4 \tab 49 \tab 11 \tab 2 \tab 5 \tab 3 \tab 20 \tab 15 \tab - \tab 2 \tab 10 \tab 5 \tab 94 \tab 1 \tab 18 \tab 2 \tab 1 \tab 2 \tab 13 \tab - \tab 1 \tab 8 \tab 8 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1124 \tab 2 \tab - \tab 1 \tab 5 \tab - \tab 14 \tab 2 \tab 11 \tab 10 \tab 6 \tab - \tab 33 \tab - \tab 2 \tab 23 \tab 1 \tab - \tab 6 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 44 \tab 13 \tab 24 \tab 92 \tab 11 \tab 6 \tab 12 \tab 18 \tab 8 \tab 2 \tab - \tab 12 \tab 10 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 5 \tab 4 \tab - \tab 9 \tab 8 \tab 3 \tab 6 \tab 10 \tab 7 \tab 6 \tab 100 \tab 13 \tab - \tab - \tab 8 \tab 21 \tab 16 \tab 6 \tab 2 \tab - \tab 15 \tab - \tab 4 \tab - \tab 6 \tab - \tab 3 \tab 24 \tab 12 \tab 7 \tab 5 \tab 135 \tab 41 \tab 5 \tab 15 \tab 1 \tab 4 \tab 6 \tab 4 \tab 28 \tab 19 \tab - \tab 11 \tab 10 \tab 7 \tab 58 \tab 71 \tab 33 \tab 2 \tab 7 \tab 16 \tab 1 \tab - \tab 7 \tab 2 \tab 86 \tab 1 \tab - \tab 1 \tab 25 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab 4 \tab 6 \tab 2 \tab 13 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 39 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 3 \tab 3 \tab 4 \tab 39 \tab 14 \tab 2 \tab 4 \tab 6 \tab 8 \tab 7 \tab - \tab 5 \tab 9 \tab 6 \tab 89 \tab 2 \tab 20 \tab 2 \tab 1 \tab 1 \tab 23 \tab 1 \tab 2 \tab 6 \tab 8 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 6 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2690 \tab 7 \tab - \tab 6 \tab 8 \tab - \tab 26 \tab 2 \tab 15 \tab 5 \tab 9 \tab 4 \tab 40 \tab 1 \tab 1 \tab 20 \tab 1 \tab 1 \tab 10 \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab 54 \tab 16 \tab 23 \tab 85 \tab 15 \tab 6 \tab 11 \tab 24 \tab 2 \tab 2 \tab - \tab 7 \tab 16 \tab 6 \tab 8 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 12 \tab 12 \tab 4 \tab 3 \tab 5 \tab 4 \tab 7 \tab 103 \tab 17 \tab 1 \tab - \tab 4 \tab 8 \tab 17 \tab 3 \tab 1 \tab 2 \tab 14 \tab - \tab 1 \tab - \tab 10 \tab 1 \tab 1 \tab 20 \tab 11 \tab 6 \tab 3 \tab 145 \tab 44 \tab 2 \tab 12 \tab 1 \tab 5 \tab 5 \tab 9 \tab 35 \tab 11 \tab 1 \tab 9 \tab 9 \tab 9 \tab 58 \tab 62 \tab 32 \tab - \tab 7 \tab 22 \tab - \tab 6 \tab 1 \tab 2 \tab 72 \tab 2 \tab - \tab - \tab 52 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 9 \tab 7 \tab 2 \tab 13 \tab - \tab 2 \tab 1 \tab 5 \tab - \tab 48 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 3 \tab 6 \tab 56 \tab 16 \tab 2 \tab 7 \tab 6 \tab 8 \tab 14 \tab - \tab 4 \tab 13 \tab 8 \tab 110 \tab 8 \tab 23 \tab 3 \tab 2 \tab - \tab 25 \tab - \tab 2 \tab 7 \tab 14 \tab - \tab 7 \tab 8 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 730 \tab 2 \tab - \tab 2 \tab 10 \tab - \tab 24 \tab 2 \tab 14 \tab 6 \tab 5 \tab 2 \tab 20 \tab - \tab 2 \tab 23 \tab 2 \tab - \tab 4 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 40 \tab 4 \tab 17 \tab 49 \tab 10 \tab 4 \tab 2 \tab 14 \tab 6 \tab 1 \tab - \tab 6 \tab 15 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 11 \tab 1 \tab 4 \tab 1 \tab 6 \tab 8 \tab 5 \tab 69 \tab 14 \tab - \tab - \tab 4 \tab 8 \tab 13 \tab 6 \tab 2 \tab 3 \tab 12 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 13 \tab 8 \tab 7 \tab - \tab 102 \tab 31 \tab 2 \tab 10 \tab 1 \tab 1 \tab 10 \tab 7 \tab 30 \tab 13 \tab 1 \tab 7 \tab 8 \tab 9 \tab 54 \tab 46 \tab 21 \tab 1 \tab 12 \tab 20 \tab 1 \tab 3 \tab 3 \tab - \tab 60 \tab - \tab - \tab - \tab 21 \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab 3 \tab 3 \tab 16 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 17 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 4 \tab 4 \tab 31 \tab 11 \tab 6 \tab 8 \tab 5 \tab 8 \tab 11 \tab - \tab 2 \tab 8 \tab 5 \tab 80 \tab 1 \tab 9 \tab - \tab - \tab 1 \tab 14 \tab 1 \tab - \tab 9 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 677 \tab 5 \tab - \tab 2 \tab 12 \tab - \tab 25 \tab 2 \tab 13 \tab 5 \tab 7 \tab 4 \tab 32 \tab - \tab 2 \tab 8 \tab - \tab - \tab 5 \tab - \tab - \tab 2 \tab 4 \tab 6 \tab 1 \tab 3 \tab - \tab 48 \tab 6 \tab 27 \tab 59 \tab 6 \tab 8 \tab 6 \tab 23 \tab 7 \tab 1 \tab 1 \tab 11 \tab 15 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 8 \tab 4 \tab 7 \tab 2 \tab 7 \tab 2 \tab 5 \tab 77 \tab 6 \tab 1 \tab 1 \tab 5 \tab 13 \tab 15 \tab 4 \tab 2 \tab 1 \tab 6 \tab - \tab 1 \tab 1 \tab 7 \tab 1 \tab - \tab 15 \tab 8 \tab 5 \tab 4 \tab 102 \tab 34 \tab 1 \tab 12 \tab 6 \tab 2 \tab 7 \tab 6 \tab 24 \tab 9 \tab - \tab 11 \tab 12 \tab 13 \tab 43 \tab 65 \tab 20 \tab 2 \tab 5 \tab 25 \tab - \tab - \tab 3 \tab 1 \tab 52 \tab 1 \tab - \tab - \tab 30 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 7 \tab 3 \tab 1 \tab 11 \tab 3 \tab 1 \tab - \tab 6 \tab 1 \tab 33 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 39 \tab 7 \tab 2 \tab 4 \tab 5 \tab 10 \tab 5 \tab - \tab 1 \tab 3 \tab 7 \tab 66 \tab 2 \tab 11 \tab 4 \tab - \tab 2 \tab 10 \tab - \tab - \tab 11 \tab 4 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 647 \tab 24 \tab - \tab 4 \tab 3 \tab 1 \tab 14 \tab - \tab 12 \tab 12 \tab 2 \tab 1 \tab 37 \tab 1 \tab 1 \tab 16 \tab - \tab - \tab 8 \tab 4 \tab 1 \tab 2 \tab 2 \tab 5 \tab - \tab 2 \tab - \tab 44 \tab 7 \tab 21 \tab 46 \tab 8 \tab 7 \tab 4 \tab 17 \tab 5 \tab 2 \tab 2 \tab 5 \tab 14 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 12 \tab 3 \tab 4 \tab 8 \tab 3 \tab 3 \tab 4 \tab 47 \tab 10 \tab - \tab 1 \tab 6 \tab 7 \tab 16 \tab 3 \tab 3 \tab 3 \tab 17 \tab - \tab - \tab - \tab 7 \tab - \tab - \tab 17 \tab 5 \tab 1 \tab 2 \tab 85 \tab 35 \tab - \tab 11 \tab 3 \tab 2 \tab 5 \tab 6 \tab 21 \tab 17 \tab - \tab 8 \tab 16 \tab 8 \tab 40 \tab 52 \tab 24 \tab 1 \tab 4 \tab 26 \tab 1 \tab - \tab 4 \tab 4 \tab 56 \tab 2 \tab - \tab - \tab 14 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 6 \tab 4 \tab 4 \tab 14 \tab - \tab - \tab 1 \tab 5 \tab - \tab 44 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 37 \tab 14 \tab 4 \tab 4 \tab 5 \tab 6 \tab 15 \tab - \tab 6 \tab 5 \tab 1 \tab 74 \tab 2 \tab 25 \tab 1 \tab 3 \tab - \tab 11 \tab - \tab - \tab 5 \tab 13 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 866 \tab 1 \tab 1 \tab 3 \tab 8 \tab 1 \tab 15 \tab - \tab 10 \tab 13 \tab 6 \tab 5 \tab 46 \tab 1 \tab 5 \tab 19 \tab - \tab 1 \tab 7 \tab 3 \tab - \tab 1 \tab - \tab 7 \tab - \tab - \tab 1 \tab 61 \tab 15 \tab 30 \tab 51 \tab 14 \tab 7 \tab 11 \tab 24 \tab 4 \tab 3 \tab - \tab 13 \tab 18 \tab 5 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 7 \tab 9 \tab 3 \tab 3 \tab 10 \tab 4 \tab 5 \tab 93 \tab 15 \tab 1 \tab - \tab 8 \tab 18 \tab 10 \tab 3 \tab 2 \tab 5 \tab 10 \tab 4 \tab 1 \tab - \tab 6 \tab 1 \tab 1 \tab 21 \tab 7 \tab 7 \tab 2 \tab 144 \tab 34 \tab 6 \tab 10 \tab 4 \tab 5 \tab 6 \tab 17 \tab 41 \tab 16 \tab - \tab 21 \tab 8 \tab 11 \tab 49 \tab 68 \tab 33 \tab 3 \tab 12 \tab 28 \tab - \tab 3 \tab 4 \tab 5 \tab 57 \tab - \tab - \tab - \tab 20 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 5 \tab 4 \tab 1 \tab 9 \tab 1 \tab 2 \tab 3 \tab 5 \tab - \tab 60 \tab 1 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 36 \tab 15 \tab 1 \tab 3 \tab 2 \tab 5 \tab 10 \tab - \tab 5 \tab 11 \tab 7 \tab 78 \tab 1 \tab 22 \tab 2 \tab 1 \tab - \tab 17 \tab 1 \tab - \tab 9 \tab 8 \tab - \tab 4 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 875 \tab 4 \tab 1 \tab - \tab 10 \tab - \tab 16 \tab 2 \tab 25 \tab 10 \tab 12 \tab 5 \tab 44 \tab 1 \tab 2 \tab 26 \tab - \tab 1 \tab 7 \tab 2 \tab 1 \tab 3 \tab 1 \tab 8 \tab 1 \tab - \tab 2 \tab 108 \tab 16 \tab 26 \tab 71 \tab 17 \tab 10 \tab 20 \tab 20 \tab 3 \tab 1 \tab - \tab 12 \tab 26 \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 4 \tab - \tab 10 \tab 9 \tab 9 \tab 5 \tab 11 \tab 5 \tab 10 \tab 92 \tab 21 \tab 2 \tab - \tab 6 \tab 10 \tab 27 \tab 3 \tab 2 \tab 5 \tab 25 \tab 2 \tab 2 \tab 1 \tab 16 \tab 1 \tab 1 \tab 30 \tab 12 \tab 3 \tab 10 \tab 125 \tab 42 \tab 2 \tab 9 \tab 3 \tab 8 \tab 10 \tab 8 \tab 43 \tab 17 \tab 1 \tab 16 \tab 6 \tab 6 \tab 80 \tab 74 \tab 40 \tab - \tab 18 \tab 25 \tab - \tab 2 \tab 6 \tab 5 \tab 61 \tab 2 \tab - \tab 1 \tab 24 \tab 4 \tab - \tab 4 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 7 \tab 10 \tab 4 \tab 20 \tab - \tab 1 \tab 1 \tab 6 \tab - \tab 46 \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab 3 \tab - \tab 37 \tab 16 \tab 4 \tab 4 \tab 5 \tab 13 \tab 10 \tab - \tab 3 \tab 15 \tab 5 \tab 128 \tab 1 \tab 26 \tab 7 \tab 2 \tab - \tab 21 \tab 2 \tab 2 \tab 6 \tab 6 \tab - \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 691 \tab 1 \tab - \tab 5 \tab 4 \tab - \tab 10 \tab 4 \tab 16 \tab 17 \tab 17 \tab 1 \tab 38 \tab 1 \tab 1 \tab 18 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 1 \tab 2 \tab 5 \tab - \tab 2 \tab - \tab 66 \tab 12 \tab 30 \tab 47 \tab 10 \tab 6 \tab 22 \tab 32 \tab 5 \tab 2 \tab 1 \tab 4 \tab 18 \tab 3 \tab 4 \tab 2 \tab - \tab - \tab - \tab 3 \tab 2 \tab 3 \tab - \tab 11 \tab 7 \tab 4 \tab 7 \tab 7 \tab 3 \tab 9 \tab 85 \tab 12 \tab 5 \tab - \tab 6 \tab 6 \tab 11 \tab 6 \tab 3 \tab 6 \tab 16 \tab - \tab - \tab 1 \tab 17 \tab 2 \tab 1 \tab 14 \tab 10 \tab 4 \tab 9 \tab 127 \tab 35 \tab 2 \tab 12 \tab 3 \tab 7 \tab 14 \tab 10 \tab 50 \tab 16 \tab 3 \tab 19 \tab 7 \tab 11 \tab 58 \tab 61 \tab 40 \tab 3 \tab 4 \tab 27 \tab - \tab 1 \tab 6 \tab 4 \tab 65 \tab 2 \tab 1 \tab 2 \tab 7 \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 5 \tab 3 \tab 4 \tab 19 \tab - \tab - \tab 2 \tab 10 \tab - \tab 43 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 21 \tab 10 \tab 4 \tab 3 \tab 5 \tab 7 \tab 16 \tab - \tab 5 \tab 17 \tab 8 \tab 70 \tab 1 \tab 23 \tab 3 \tab - \tab - \tab 14 \tab 1 \tab 2 \tab 8 \tab 3 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 51301 \tab 51 \tab 3 \tab 42 \tab 109 \tab 4 \tab 274 \tab 22 \tab 188 \tab 113 \tab 89 \tab 34 \tab 475 \tab 7 \tab 26 \tab 219 \tab 6 \tab 10 \tab 69 \tab 15 \tab 7 \tab 18 \tab 21 \tab 49 \tab 4 \tab 15 \tab 9 \tab 737 \tab 121 \tab 298 \tab 833 \tab 136 \tab 76 \tab 127 \tab 242 \tab 52 \tab 20 \tab 6 \tab 104 \tab 207 \tab 48 \tab 35 \tab 7 \tab 5 \tab 2 \tab 1 \tab 9 \tab 24 \tab 26 \tab 2 \tab 143 \tab 75 \tab 60 \tab 62 \tab 87 \tab 57 \tab 72 \tab 1148 \tab 174 \tab 18 \tab 6 \tab 67 \tab 147 \tab 180 \tab 48 \tab 24 \tab 42 \tab 162 \tab 10 \tab 15 \tab 4 \tab 104 \tab 9 \tab 15 \tab 260 \tab 117 \tab 61 \tab 52 \tab 1700 \tab 531 \tab 28 \tab 159 \tab 31 \tab 58 \tab 113 \tab 103 \tab 459 \tab 166 \tab 10 \tab 153 \tab 141 \tab 112 \tab 706 \tab 810 \tab 362 \tab 16 \tab 87 \tab 314 \tab 5 \tab 22 \tab 61 \tab 27 \tab 869 \tab 13 \tab 3 \tab 6 \tab 343 \tab 31 \tab 4 \tab 17 \tab 12 \tab 12 \tab 20 \tab 7 \tab 11 \tab 3 \tab 8 \tab 13 \tab 16 \tab 68 \tab 56 \tab 32 \tab 214 \tab 12 \tab 11 \tab 17 \tab 95 \tab 4 \tab 553 \tab 14 \tab 15 \tab 3 \tab 18 \tab 41 \tab 11 \tab 60 \tab 44 \tab 605 \tab 196 \tab 50 \tab 69 \tab 66 \tab 127 \tab 155 \tab 1 \tab 46 \tab 124 \tab 81 \tab 1189 \tab 26 \tab 238 \tab 36 \tab 14 \tab 9 \tab 215 \tab 7 \tab 19 \tab 112 \tab 98 \tab 8 \tab 29 \tab 39 \tab 5 \tab 9 \tab 3 \tab 20 \tab 9 \tab 14 \tab 23 \tab 6 \tab 8 \tab 7 \tab 8 \tab 1 \tab 4 \tab 11 \tab 9 \tab 17 \tab 12 \tab 7 \tab 17 \tab 9 \tab 9 \tab 3 \tab 1 \tab 9 \tab 9 \tab 13 \tab 10 \tab 10 \tab 7 \tab 3 \tab 4 \tab 9 \tab 11 \tab 8 \tab 17 \tab 5 \tab 2 \tab 4 \tab 1 \tab 3 \tab 1 \tab 4 \tab 7 \tab 4 \tab 12 \tab 6 \tab 6 \tab 3 \tab 8 \tab 8 \tab 3 \tab 6 \tab 2 \tab 11 \tab 3 \tab 4 \tab 2 \tab 2 \tab 5 \tab 2 \tab 1 \tab 2 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{9997} uncodeable
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
