% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{maoccindv}
\alias{maoccindv}
\title{Mothers occ/ind verbatim retrieved}
\source{
General Social Survey https://gss.norc.org
}
\description{
maoccindv
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab no verbatims for ma \tab verbatims for ma \tab Total \cr
   1972 \tab 911 \tab 702 \tab - \tab 1613 \cr
   1973 \tab 270 \tab 1234 \tab - \tab 1504 \cr
   1974 \tab 267 \tab 1217 \tab - \tab 1484 \cr
   1975 \tab 514 \tab 976 \tab - \tab 1490 \cr
   1976 \tab 269 \tab 1230 \tab - \tab 1499 \cr
   1977 \tab 713 \tab 817 \tab - \tab 1530 \cr
   1978 \tab 14 \tab 1518 \tab - \tab 1532 \cr
   1980 \tab 175 \tab 1293 \tab - \tab 1468 \cr
   1982 \tab 164 \tab 1696 \tab - \tab 1860 \cr
   1983 \tab 124 \tab 1475 \tab - \tab 1599 \cr
   1984 \tab 15 \tab 1458 \tab - \tab 1473 \cr
   1985 \tab 238 \tab 1296 \tab - \tab 1534 \cr
   1986 \tab 45 \tab 1425 \tab - \tab 1470 \cr
   1987 \tab 308 \tab 1511 \tab - \tab 1819 \cr
   1988 \tab 21 \tab 1460 \tab - \tab 1481 \cr
   1989 \tab 117 \tab 1420 \tab - \tab 1537 \cr
   1990 \tab 14 \tab 1358 \tab - \tab 1372 \cr
   1991 \tab 18 \tab 1499 \tab - \tab 1517 \cr
   1993 \tab 61 \tab 1545 \tab - \tab 1606 \cr
   1994 \tab 54 \tab 1228 \tab 1710 \tab 2992 \cr
   1996 \tab 21 \tab 1101 \tab 1782 \tab 2904 \cr
   1998 \tab 31 \tab 1056 \tab 1745 \tab 2832 \cr
   2000 \tab - \tab 1172 \tab 1645 \tab 2817 \cr
   2002 \tab - \tab 1123 \tab 1642 \tab 2765 \cr
   2004 \tab - \tab 1104 \tab 1708 \tab 2812 \cr
   2006 \tab - \tab 2674 \tab 1836 \tab 4510 \cr
   2008 \tab - \tab 709 \tab 1314 \tab 2023 \cr
   2010 \tab - \tab 666 \tab 1378 \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab 3544 \cr
   Total \tab 21667 \tab 35963 \tab 14760 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} no verbatims for ma
\item \code{1} verbatims for ma
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
