% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_39.R
\name{mapa}
\alias{mapa}
\title{Contrast between mother and father}
\source{
General Social Survey https://gss.norc.org
}
\description{
mapa
}
\details{
Question 114. There are many different ways of picturing God. We'd like to know the kinds of images you are most likely to associate with God. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of God between the two contrasting images? a. 1. Mother 7. Father
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab father \tab mother \tab no answer \tab don't know \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 484 \tab 16 \tab 30 \tab 257 \tab 86 \tab 80 \tab 429 \tab 32 \tab 59 \tab - \tab 1473 \cr
   1985 \tab - \tab 15 \tab 26 \tab 353 \tab 142 \tab 140 \tab 732 \tab 50 \tab 58 \tab 18 \tab 1534 \cr
   1986 \tab - \tab 15 \tab 24 \tab 347 \tab 109 \tab 123 \tab 735 \tab 62 \tab 47 \tab 8 \tab 1470 \cr
   1987 \tab - \tab 23 \tab 22 \tab 454 \tab 126 \tab 133 \tab 895 \tab 64 \tab 81 \tab 21 \tab 1819 \cr
   1988 \tab - \tab 26 \tab 26 \tab 372 \tab 105 \tab 104 \tab 720 \tab 41 \tab 59 \tab 28 \tab 1481 \cr
   1989 \tab 531 \tab 10 \tab 12 \tab 257 \tab 85 \tab 93 \tab 471 \tab 33 \tab 41 \tab 4 \tab 1537 \cr
   1990 \tab 444 \tab 14 \tab 14 \tab 229 \tab 77 \tab 91 \tab 437 \tab 20 \tab 38 \tab 8 \tab 1372 \cr
   1991 \tab 158 \tab 26 \tab 52 \tab 325 \tab 114 \tab 116 \tab 522 \tab 42 \tab 162 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 12 \tab 29 \tab 251 \tab 85 \tab 104 \tab 506 \tab 33 \tab 54 \tab 6 \tab 1606 \cr
   1994 \tab 1015 \tab 28 \tab 33 \tab 504 \tab 158 \tab 174 \tab 940 \tab 48 \tab 72 \tab 20 \tab 2992 \cr
   1996 \tab 1913 \tab 14 \tab 22 \tab 231 \tab 93 \tab 82 \tab 467 \tab 36 \tab 35 \tab 11 \tab 2904 \cr
   1998 \tab 1387 \tab 16 \tab 30 \tab 354 \tab 119 \tab 139 \tab 681 \tab 42 \tab 50 \tab 14 \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 658 \tab 15 \tab 43 \tab 340 \tab 102 \tab 118 \tab 599 \tab 50 \tab 32 \tab 66 \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 54946 \tab 230 \tab 363 \tab 4274 \tab 1401 \tab 1497 \tab 8134 \tab 553 \tab 788 \tab 204 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} mother
\item \code{7} father
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
