% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_09.R
\name{mapres10}
\alias{mapres10}
\title{mapres10}
\source{
General Social Survey https://gss.norc.org
}
\description{
Mothers occupational prestige score (2010)
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 16 \tab 18 \tab 20 \tab 21 \tab 22 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 68 \tab 69 \tab 72 \tab 73 \tab 74 \tab 80 \tab 23 \tab 67 \tab 70 \tab 71 \tab 17 \tab 75 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1351 \tab 5 \tab 2 \tab 1 \tab 18 \tab 7 \tab 36 \tab 148 \tab 16 \tab 42 \tab 46 \tab 9 \tab 9 \tab 152 \tab 101 \tab 59 \tab 8 \tab 105 \tab 32 \tab 13 \tab 71 \tab 52 \tab 10 \tab 4 \tab 31 \tab 4 \tab 7 \tab 66 \tab 41 \tab 147 \tab 96 \tab 18 \tab 22 \tab 8 \tab 5 \tab 15 \tab 4 \tab 12 \tab 2 \tab 3 \tab 3 \tab 4 \tab 20 \tab 71 \tab 2 \tab 1 \tab 71 \tab 2 \tab 3 \tab 1 \tab 26 \tab 3 \tab 2 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1158 \tab 14 \tab 1 \tab 1 \tab 13 \tab 7 \tab 37 \tab 137 \tab 24 \tab 35 \tab 45 \tab 9 \tab 8 \tab 146 \tab 97 \tab 63 \tab 8 \tab 113 \tab 38 \tab 21 \tab 72 \tab 47 \tab 14 \tab 6 \tab 22 \tab 8 \tab 12 \tab 82 \tab 43 \tab 159 \tab 101 \tab 17 \tab 23 \tab 16 \tab 6 \tab 14 \tab 12 \tab 12 \tab 3 \tab - \tab 4 \tab 8 \tab 21 \tab 80 \tab 2 \tab 3 \tab 85 \tab 2 \tab 3 \tab - \tab 23 \tab 9 \tab 2 \tab 8 \tab 3 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 2904 \cr
   1998 \tab 1213 \tab 5 \tab 1 \tab - \tab 14 \tab 16 \tab 31 \tab 130 \tab 9 \tab 27 \tab 51 \tab 15 \tab 4 \tab 127 \tab 99 \tab 55 \tab 6 \tab 134 \tab 33 \tab 13 \tab 73 \tab 51 \tab 16 \tab 3 \tab 26 \tab 3 \tab 6 \tab 57 \tab 28 \tab 132 \tab 98 \tab 19 \tab 14 \tab 6 \tab 6 \tab 20 \tab 15 \tab 19 \tab 4 \tab - \tab - \tab 13 \tab 15 \tab 82 \tab 1 \tab 2 \tab 92 \tab - \tab 1 \tab - \tab 19 \tab 6 \tab 5 \tab 5 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 6 \tab - \tab 2832 \cr
   2000 \tab 1184 \tab 5 \tab 3 \tab 5 \tab 20 \tab 19 \tab 34 \tab 124 \tab 12 \tab 27 \tab 41 \tab 10 \tab 9 \tab 125 \tab 84 \tab 66 \tab 10 \tab 120 \tab 33 \tab 9 \tab 72 \tab 49 \tab 12 \tab 2 \tab 37 \tab 4 \tab 7 \tab 61 \tab 29 \tab 165 \tab 97 \tab 20 \tab 18 \tab 10 \tab 10 \tab 24 \tab 13 \tab 18 \tab 3 \tab 1 \tab 1 \tab 6 \tab 12 \tab 70 \tab 1 \tab 3 \tab 80 \tab 2 \tab - \tab - \tab 20 \tab 5 \tab 2 \tab 9 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 2817 \cr
   2002 \tab 1139 \tab 9 \tab 2 \tab 1 \tab 17 \tab 10 \tab 40 \tab 120 \tab 10 \tab 38 \tab 34 \tab 7 \tab 2 \tab 151 \tab 92 \tab 61 \tab 9 \tab 111 \tab 34 \tab 7 \tab 75 \tab 43 \tab 16 \tab 1 \tab 20 \tab 3 \tab 7 \tab 50 \tab 39 \tab 180 \tab 131 \tab 16 \tab 19 \tab 9 \tab 4 \tab 25 \tab 10 \tab 15 \tab 3 \tab 1 \tab 4 \tab 6 \tab 18 \tab 69 \tab 2 \tab 7 \tab 54 \tab 4 \tab 1 \tab 1 \tab 12 \tab 5 \tab 3 \tab 6 \tab 4 \tab 2 \tab - \tab 3 \tab - \tab 3 \tab - \tab 2765 \cr
   2004 \tab 1124 \tab 7 \tab 1 \tab 2 \tab 21 \tab 7 \tab 35 \tab 119 \tab 12 \tab 34 \tab 34 \tab 8 \tab 5 \tab 164 \tab 76 \tab 63 \tab 15 \tab 114 \tab 25 \tab 18 \tab 70 \tab 55 \tab 16 \tab 9 \tab 22 \tab 3 \tab 11 \tab 54 \tab 24 \tab 150 \tab 106 \tab 29 \tab 11 \tab 15 \tab 7 \tab 22 \tab 11 \tab 21 \tab 5 \tab 1 \tab 4 \tab 11 \tab 21 \tab 103 \tab 3 \tab 5 \tab 72 \tab 3 \tab 3 \tab - \tab 30 \tab 2 \tab 2 \tab 13 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 4 \tab 2812 \cr
   2006 \tab 2690 \tab 9 \tab 1 \tab 8 \tab 16 \tab 9 \tab 37 \tab 145 \tab 13 \tab 43 \tab 45 \tab 4 \tab 8 \tab 144 \tab 88 \tab 63 \tab 11 \tab 150 \tab 28 \tab 19 \tab 77 \tab 46 \tab 13 \tab 5 \tab 21 \tab 3 \tab 10 \tab 63 \tab 43 \tab 158 \tab 127 \tab 15 \tab 24 \tab 10 \tab 10 \tab 37 \tab 14 \tab 14 \tab 2 \tab 1 \tab - \tab 7 \tab 34 \tab 100 \tab 5 \tab 6 \tab 77 \tab 6 \tab 2 \tab - \tab 14 \tab 8 \tab 4 \tab 16 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 4510 \cr
   2008 \tab 730 \tab 9 \tab 1 \tab 1 \tab 17 \tab 11 \tab 32 \tab 98 \tab 9 \tab 27 \tab 38 \tab 8 \tab 6 \tab 99 \tab 52 \tab 55 \tab 5 \tab 81 \tab 24 \tab 12 \tab 42 \tab 49 \tab 5 \tab 11 \tab 22 \tab 8 \tab 7 \tab 54 \tab 29 \tab 113 \tab 100 \tab 15 \tab 17 \tab 8 \tab 6 \tab 20 \tab 11 \tab 7 \tab 2 \tab 1 \tab - \tab 6 \tab 20 \tab 55 \tab 5 \tab 4 \tab 58 \tab 3 \tab 2 \tab - \tab 8 \tab 3 \tab 4 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 2023 \cr
   2010 \tab 677 \tab 12 \tab 2 \tab 2 \tab 19 \tab 12 \tab 26 \tab 85 \tab 5 \tab 20 \tab 35 \tab 5 \tab 7 \tab 121 \tab 68 \tab 45 \tab 9 \tab 89 \tab 21 \tab 13 \tab 56 \tab 44 \tab 15 \tab 2 \tab 20 \tab 4 \tab 6 \tab 44 \tab 37 \tab 118 \tab 84 \tab 16 \tab 14 \tab 13 \tab 6 \tab 18 \tab 11 \tab 18 \tab 5 \tab 2 \tab 1 \tab 3 \tab 25 \tab 70 \tab 3 \tab 5 \tab 77 \tab 2 \tab 3 \tab 1 \tab 14 \tab 5 \tab 4 \tab 6 \tab 6 \tab 2 \tab 2 \tab 2 \tab - \tab 5 \tab 2 \tab 2044 \cr
   2012 \tab 647 \tab 8 \tab 2 \tab 2 \tab 8 \tab 8 \tab 44 \tab 99 \tab 10 \tab 18 \tab 28 \tab 5 \tab 5 \tab 125 \tab 67 \tab 46 \tab 8 \tab 100 \tab 29 \tab 6 \tab 61 \tab 48 \tab 7 \tab 5 \tab 19 \tab 3 \tab 14 \tab 51 \tab 27 \tab 94 \tab 59 \tab 11 \tab 13 \tab 14 \tab 7 \tab 13 \tab 6 \tab 18 \tab 3 \tab - \tab 2 \tab 9 \tab 18 \tab 56 \tab 2 \tab 11 \tab 72 \tab 5 \tab 2 \tab - \tab 13 \tab 4 \tab 22 \tab 7 \tab 5 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab 1974 \cr
   2014 \tab 866 \tab 11 \tab 2 \tab 1 \tab 18 \tab 12 \tab 43 \tab 96 \tab 13 \tab 24 \tab 48 \tab 3 \tab 9 \tab 115 \tab 74 \tab 57 \tab 12 \tab 137 \tab 37 \tab 17 \tab 93 \tab 42 \tab 8 \tab 5 \tab 22 \tab 14 \tab 6 \tab 48 \tab 29 \tab 162 \tab 115 \tab 24 \tab 19 \tab 10 \tab 12 \tab 28 \tab 15 \tab 13 \tab 5 \tab 2 \tab 2 \tab 9 \tab 25 \tab 62 \tab 3 \tab 5 \tab 99 \tab 4 \tab 1 \tab - \tab 23 \tab 7 \tab 2 \tab 15 \tab 7 \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 2538 \cr
   2016 \tab 875 \tab 6 \tab 1 \tab 1 \tab 17 \tab 10 \tab 46 \tab 147 \tab 12 \tab 38 \tab 53 \tab 15 \tab 3 \tab 145 \tab 83 \tab 84 \tab 7 \tab 138 \tab 36 \tab 17 \tab 94 \tab 57 \tab 18 \tab 7 \tab 25 \tab 6 \tab 15 \tab 75 \tab 56 \tab 140 \tab 140 \tab 19 \tab 18 \tab 21 \tab 15 \tab 42 \tab 18 \tab 22 \tab 6 \tab - \tab 5 \tab 10 \tab 25 \tab 89 \tab 4 \tab 4 \tab 135 \tab 4 \tab 1 \tab 1 \tab 15 \tab 7 \tab 2 \tab 16 \tab 8 \tab 2 \tab 2 \tab 2 \tab 1 \tab 6 \tab - \tab 2867 \cr
   2018 \tab 691 \tab 10 \tab 1 \tab 1 \tab 24 \tab 8 \tab 51 \tab 89 \tab 7 \tab 22 \tab 57 \tab 7 \tab 6 \tab 122 \tab 47 \tab 66 \tab 8 \tab 124 \tab 29 \tab 11 \tab 87 \tab 45 \tab 23 \tab 3 \tab 28 \tab 9 \tab 11 \tab 56 \tab 44 \tab 144 \tab 108 \tab 14 \tab 25 \tab 12 \tab 9 \tab 34 \tab 12 \tab 20 \tab 7 \tab 1 \tab 2 \tab 13 \tab 39 \tab 58 \tab 1 \tab 4 \tab 103 \tab 1 \tab 4 \tab - \tab 12 \tab 7 \tab 2 \tab 12 \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab 8 \tab - \tab 2348 \cr
   2021 \tab 1265 \tab 16 \tab 4 \tab 4 \tab 38 \tab 12 \tab 39 \tab 137 \tab 19 \tab 35 \tab 65 \tab 8 \tab 10 \tab 241 \tab 160 \tab 78 \tab 20 \tab 184 \tab 55 \tab 27 \tab 117 \tab 73 \tab 25 \tab 6 \tab 31 \tab 12 \tab 17 \tab 124 \tab 53 \tab 258 \tab 160 \tab 35 \tab 25 \tab 32 \tab 5 \tab 48 \tab 35 \tab 25 \tab 11 \tab 5 \tab 3 \tab 18 \tab 35 \tab 173 \tab 7 \tab 9 \tab 152 \tab 7 \tab 5 \tab - \tab 51 \tab 12 \tab 2 \tab 13 \tab 8 \tab 1 \tab 6 \tab 3 \tab 3 \tab 8 \tab 2 \tab 4032 \cr
   2022 \tab 1040 \tab 9 \tab 3 \tab 2 \tab 26 \tab 9 \tab 56 \tab 156 \tab 15 \tab 31 \tab 62 \tab 4 \tab 6 \tab 214 \tab 116 \tab 71 \tab 13 \tab 167 \tab 47 \tab 18 \tab 109 \tab 79 \tab 24 \tab 6 \tab 27 \tab 10 \tab 14 \tab 96 \tab 45 \tab 202 \tab 162 \tab 32 \tab 33 \tab 19 \tab 15 \tab 46 \tab 30 \tab 21 \tab 10 \tab 2 \tab 5 \tab 16 \tab 42 \tab 140 \tab 10 \tab 16 \tab 156 \tab 11 \tab 6 \tab - \tab 43 \tab 11 \tab 5 \tab 11 \tab 9 \tab - \tab 4 \tab 5 \tab 1 \tab 5 \tab 1 \tab 3544 \cr
   Total \tab 46038 \tab 135 \tab 27 \tab 32 \tab 286 \tab 157 \tab 587 \tab 1830 \tab 186 \tab 461 \tab 682 \tab 117 \tab 97 \tab 2191 \tab 1304 \tab 932 \tab 149 \tab 1867 \tab 501 \tab 221 \tab 1169 \tab 780 \tab 222 \tab 75 \tab 373 \tab 94 \tab 150 \tab 981 \tab 567 \tab 2322 \tab 1684 \tab 300 \tab 295 \tab 203 \tab 123 \tab 406 \tab 217 \tab 255 \tab 71 \tab 20 \tab 36 \tab 139 \tab 370 \tab 1278 \tab 51 \tab 85 \tab 1383 \tab 56 \tab 37 \tab 4 \tab 323 \tab 94 \tab 63 \tab 145 \tab 66 \tab 19 \tab 21 \tab 21 \tab 21 \tab 59 \tab 12 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
