% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_09.R
\name{mapres105plus}
\alias{mapres105plus}
\title{mapres105plus}
\source{
General Social Survey https://gss.norc.org
}
\description{
Mothers occupational prestige score using threshold method (2010)
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 42 \tab 43 \tab 44 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 58 \tab 59 \tab 6 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 7 \tab 70 \tab 72 \tab 73 \tab 74 \tab 75 \tab 76 \tab 78 \tab 79 \tab 8 \tab 80 \tab 82 \tab 83 \tab 84 \tab 85 \tab 86 \tab 87 \tab 88 \tab 89 \tab 9 \tab 90 \tab 91 \tab 92 \tab 93 \tab 94 \tab 95 \tab 71 \tab 77 \tab 81 \tab 97 \tab 57 \tab 41 \tab 45 \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1351 \tab 15 \tab 9 \tab 58 \tab 10 \tab 116 \tab 21 \tab 99 \tab 2 \tab 105 \tab 8 \tab 9 \tab 28 \tab 69 \tab 4 \tab 50 \tab 7 \tab 58 \tab 38 \tab 66 \tab 37 \tab 1 \tab 1 \tab 11 \tab 2 \tab 3 \tab 9 \tab 1 \tab 4 \tab 21 \tab 36 \tab 13 \tab 13 \tab 1 \tab 8 \tab 22 \tab 7 \tab 56 \tab 5 \tab 5 \tab 88 \tab 3 \tab 1 \tab 2 \tab 22 \tab 143 \tab 8 \tab 18 \tab 2 \tab 2 \tab 8 \tab 5 \tab 13 \tab 3 \tab 5 \tab 10 \tab 10 \tab 1 \tab 9 \tab 11 \tab 1 \tab 1 \tab 1 \tab 11 \tab 2 \tab 2 \tab 3 \tab 2 \tab 2 \tab 7 \tab 3 \tab 1 \tab 4 \tab 7 \tab 80 \tab 3 \tab 52 \tab 24 \tab 2 \tab 7 \tab 6 \tab 18 \tab 1 \tab 1 \tab 5 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1158 \tab 6 \tab 6 \tab 63 \tab 10 \tab 114 \tab 17 \tab 99 \tab 5 \tab 102 \tab 4 \tab 13 \tab 18 \tab 74 \tab 1 \tab 48 \tab 8 \tab 61 \tab 39 \tab 76 \tab 63 \tab 2 \tab - \tab 13 \tab 4 \tab - \tab 5 \tab 3 \tab 4 \tab 19 \tab 29 \tab 14 \tab 7 \tab - \tab 8 \tab 24 \tab 11 \tab 57 \tab 7 \tab 14 \tab 91 \tab 7 \tab - \tab 4 \tab 22 \tab 158 \tab 16 \tab 22 \tab 2 \tab - \tab 7 \tab 3 \tab 15 \tab 3 \tab 8 \tab 15 \tab 3 \tab - \tab 16 \tab 15 \tab 1 \tab 2 \tab - \tab 6 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 8 \tab 5 \tab 1 \tab 6 \tab 6 \tab 89 \tab 3 \tab 59 \tab 23 \tab - \tab 6 \tab 11 \tab 28 \tab 1 \tab - \tab 14 \tab 3 \tab 4 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1213 \tab 14 \tab 9 \tab 52 \tab 19 \tab 91 \tab 14 \tab 92 \tab 2 \tab 94 \tab 12 \tab 14 \tab 11 \tab 70 \tab 1 \tab 46 \tab 5 \tab 64 \tab 44 \tab 81 \tab 48 \tab 4 \tab - \tab 16 \tab 2 \tab 4 \tab 8 \tab 1 \tab 5 \tab 13 \tab 34 \tab 18 \tab 11 \tab 1 \tab 7 \tab 23 \tab 5 \tab 41 \tab 8 \tab 5 \tab 89 \tab 5 \tab - \tab 2 \tab 15 \tab 129 \tab 10 \tab 14 \tab 4 \tab 6 \tab 5 \tab - \tab 11 \tab 5 \tab 6 \tab 15 \tab 3 \tab 1 \tab 27 \tab 13 \tab - \tab - \tab - \tab 7 \tab 3 \tab - \tab 2 \tab - \tab 4 \tab 4 \tab 8 \tab 1 \tab 6 \tab 2 \tab 90 \tab 4 \tab 64 \tab 17 \tab 1 \tab 4 \tab 4 \tab 28 \tab 2 \tab - \tab 9 \tab 5 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1184 \tab 6 \tab 10 \tab 44 \tab 20 \tab 105 \tab 14 \tab 80 \tab 1 \tab 92 \tab 7 \tab 10 \tab 18 \tab 56 \tab 1 \tab 55 \tab 5 \tab 58 \tab 42 \tab 81 \tab 40 \tab 3 \tab - \tab 18 \tab - \tab 1 \tab 9 \tab 2 \tab 2 \tab 11 \tab 36 \tab 17 \tab 13 \tab 1 \tab 8 \tab 39 \tab 6 \tab 44 \tab 4 \tab 5 \tab 85 \tab 7 \tab - \tab 2 \tab 15 \tab 155 \tab 11 \tab 17 \tab 7 \tab 5 \tab 1 \tab 5 \tab 10 \tab 5 \tab 5 \tab 18 \tab 9 \tab - \tab 21 \tab 14 \tab 7 \tab 2 \tab 1 \tab 10 \tab 2 \tab 4 \tab 1 \tab 5 \tab 4 \tab 13 \tab 3 \tab 1 \tab 6 \tab 2 \tab 73 \tab 4 \tab 60 \tab 20 \tab 1 \tab 8 \tab 3 \tab 23 \tab 2 \tab - \tab 13 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1139 \tab 4 \tab 8 \tab 60 \tab 13 \tab 98 \tab 20 \tab 95 \tab 1 \tab 93 \tab 4 \tab 10 \tab 20 \tab 50 \tab 1 \tab 49 \tab 9 \tab 68 \tab 39 \tab 76 \tab 39 \tab 5 \tab - \tab 19 \tab 3 \tab 3 \tab 9 \tab 1 \tab 3 \tab 14 \tab 31 \tab 15 \tab 10 \tab 1 \tab 10 \tab 18 \tab 4 \tab 28 \tab 5 \tab 9 \tab 113 \tab 4 \tab - \tab 2 \tab 20 \tab 177 \tab 19 \tab 10 \tab 6 \tab 4 \tab 4 \tab 5 \tab 15 \tab 5 \tab 8 \tab 10 \tab 5 \tab 1 \tab 20 \tab 10 \tab 2 \tab 1 \tab 3 \tab 17 \tab 3 \tab 1 \tab 4 \tab 6 \tab - \tab 6 \tab 6 \tab - \tab 5 \tab 1 \tab 78 \tab 4 \tab 45 \tab 11 \tab 2 \tab 7 \tab 4 \tab 14 \tab 2 \tab 2 \tab 7 \tab 7 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1124 \tab 4 \tab 9 \tab 46 \tab 11 \tab 93 \tab 22 \tab 93 \tab - \tab 113 \tab 5 \tab 16 \tab 21 \tab 41 \tab 2 \tab 52 \tab 9 \tab 57 \tab 33 \tab 81 \tab 58 \tab 1 \tab 1 \tab 20 \tab 5 \tab 2 \tab 6 \tab 1 \tab 9 \tab 14 \tab 29 \tab 8 \tab 18 \tab 2 \tab 16 \tab 18 \tab 4 \tab 39 \tab 15 \tab 7 \tab 88 \tab 3 \tab - \tab 2 \tab 13 \tab 142 \tab 16 \tab 12 \tab 5 \tab 5 \tab 1 \tab 7 \tab 5 \tab 5 \tab 7 \tab 14 \tab 11 \tab - \tab 18 \tab 23 \tab 2 \tab 4 \tab 2 \tab 9 \tab 6 \tab 2 \tab 1 \tab 5 \tab 4 \tab 6 \tab 7 \tab 1 \tab 8 \tab 8 \tab 110 \tab 5 \tab 48 \tab 33 \tab 2 \tab 10 \tab 3 \tab 27 \tab - \tab 2 \tab 14 \tab 6 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2812 \cr
   2006 \tab 2690 \tab 4 \tab 6 \tab 46 \tab 12 \tab 116 \tab 24 \tab 101 \tab 3 \tab 108 \tab 10 \tab 12 \tab 16 \tab 57 \tab 2 \tab 52 \tab 7 \tab 82 \tab 38 \tab 86 \tab 52 \tab 2 \tab - \tab 17 \tab 5 \tab 2 \tab 9 \tab 1 \tab - \tab 14 \tab 39 \tab 13 \tab 12 \tab 1 \tab 14 \tab 13 \tab 5 \tab 40 \tab 9 \tab 9 \tab 117 \tab 5 \tab 1 \tab 2 \tab 18 \tab 159 \tab 14 \tab 15 \tab 7 \tab 2 \tab 6 \tab 3 \tab 11 \tab 6 \tab 8 \tab 18 \tab 12 \tab - \tab 27 \tab 11 \tab 8 \tab 2 \tab - \tab 15 \tab 4 \tab 4 \tab 4 \tab 11 \tab 1 \tab 9 \tab 5 \tab - \tab 6 \tab 10 \tab 111 \tab 1 \tab 58 \tab 12 \tab 2 \tab 5 \tab 9 \tab 29 \tab 2 \tab - \tab 18 \tab 6 \tab 4 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 730 \tab 7 \tab 8 \tab 42 \tab 12 \tab 84 \tab 10 \tab 66 \tab 3 \tab 76 \tab 4 \tab 8 \tab 13 \tab 32 \tab 2 \tab 45 \tab 5 \tab 43 \tab 33 \tab 48 \tab 42 \tab 1 \tab - \tab 8 \tab 2 \tab 1 \tab 8 \tab 2 \tab 4 \tab 10 \tab 15 \tab 15 \tab 11 \tab 2 \tab 23 \tab 19 \tab 2 \tab 29 \tab 8 \tab 9 \tab 76 \tab 10 \tab - \tab 2 \tab 16 \tab 109 \tab 10 \tab 22 \tab 3 \tab 4 \tab 9 \tab 5 \tab 8 \tab 5 \tab 8 \tab 11 \tab 3 \tab - \tab 18 \tab 10 \tab 3 \tab 3 \tab - \tab 8 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 5 \tab 4 \tab 2 \tab 5 \tab 1 \tab 65 \tab 2 \tab 42 \tab 9 \tab 1 \tab 9 \tab 5 \tab 23 \tab 3 \tab - \tab 4 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 677 \tab 2 \tab 13 \tab 38 \tab 15 \tab 66 \tab 11 \tab 79 \tab 3 \tab 62 \tab 8 \tab 16 \tab 14 \tab 40 \tab 1 \tab 43 \tab 2 \tab 50 \tab 28 \tab 59 \tab 43 \tab - \tab - \tab 15 \tab 3 \tab 1 \tab 2 \tab - \tab 3 \tab 12 \tab 31 \tab 17 \tab 8 \tab - \tab 8 \tab 21 \tab 6 \tab 30 \tab 6 \tab 12 \tab 74 \tab 4 \tab - \tab 2 \tab 15 \tab 111 \tab 11 \tab 18 \tab 5 \tab 7 \tab 2 \tab 5 \tab 8 \tab 5 \tab 7 \tab 9 \tab 4 \tab - \tab 17 \tab 14 \tab 2 \tab 1 \tab 4 \tab 15 \tab 5 \tab 1 \tab 3 \tab 6 \tab 1 \tab 7 \tab 5 \tab - \tab 9 \tab 2 \tab 84 \tab 3 \tab 50 \tab 16 \tab 1 \tab 6 \tab 5 \tab 29 \tab - \tab 1 \tab 10 \tab 9 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 647 \tab 6 \tab 5 \tab 37 \tab 9 \tab 74 \tab 26 \tab 69 \tab 1 \tab 77 \tab 6 \tab 12 \tab 26 \tab 38 \tab - \tab 38 \tab 7 \tab 40 \tab 36 \tab 79 \tab 44 \tab 2 \tab - \tab 7 \tab 2 \tab 2 \tab 4 \tab 4 \tab 2 \tab 14 \tab 33 \tab 6 \tab 11 \tab - \tab 11 \tab 18 \tab 4 \tab 33 \tab 15 \tab 8 \tab 48 \tab 2 \tab - \tab 3 \tab 15 \tab 93 \tab 10 \tab 11 \tab 2 \tab 3 \tab 3 \tab 5 \tab 10 \tab 4 \tab 10 \tab 13 \tab 4 \tab - \tab 13 \tab 8 \tab 2 \tab 1 \tab 1 \tab 7 \tab 3 \tab 2 \tab - \tab 3 \tab 1 \tab 2 \tab 8 \tab 1 \tab 7 \tab 5 \tab 68 \tab 7 \tab 49 \tab 12 \tab 2 \tab 7 \tab 10 \tab 22 \tab 1 \tab 1 \tab 10 \tab 26 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 866 \tab - \tab 13 \tab 39 \tab 16 \tab 80 \tab 23 \tab 90 \tab 5 \tab 75 \tab 5 \tab 7 \tab 21 \tab 37 \tab 1 \tab 52 \tab 6 \tab 55 \tab 45 \tab 101 \tab 68 \tab 3 \tab - \tab 14 \tab 6 \tab 4 \tab 8 \tab 2 \tab 3 \tab 10 \tab 44 \tab 5 \tab 10 \tab 1 \tab 22 \tab 16 \tab 8 \tab 35 \tab 7 \tab 11 \tab 90 \tab 5 \tab - \tab 3 \tab 13 \tab 150 \tab 16 \tab 16 \tab 7 \tab 2 \tab 3 \tab 12 \tab 10 \tab 4 \tab 9 \tab 16 \tab 7 \tab 1 \tab 26 \tab 19 \tab 2 \tab 3 \tab 1 \tab 14 \tab 1 \tab - \tab 3 \tab 7 \tab 3 \tab 9 \tab 5 \tab - \tab 7 \tab 5 \tab 74 \tab 10 \tab 64 \tab 21 \tab 3 \tab 5 \tab 7 \tab 33 \tab 2 \tab - \tab 19 \tab 7 \tab 3 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab 875 \tab 3 \tab 12 \tab 43 \tab 16 \tab 128 \tab 28 \tab 103 \tab 1 \tab 94 \tab 9 \tab 24 \tab 19 \tab 37 \tab 3 \tab 77 \tab 8 \tab 65 \tab 45 \tab 96 \tab 63 \tab 3 \tab - \tab 25 \tab 4 \tab 5 \tab 12 \tab 2 \tab 6 \tab 20 \tab 34 \tab 9 \tab 18 \tab 1 \tab 28 \tab 23 \tab 9 \tab 42 \tab 18 \tab 6 \tab 102 \tab 5 \tab 1 \tab 2 \tab 30 \tab 141 \tab 27 \tab 24 \tab 6 \tab 6 \tab 8 \tab 8 \tab 8 \tab 6 \tab 18 \tab 21 \tab 4 \tab 1 \tab 33 \tab 14 \tab 1 \tab 2 \tab 2 \tab 18 \tab 8 \tab 4 \tab 4 \tab 7 \tab 4 \tab 6 \tab 5 \tab - \tab 13 \tab 5 \tab 94 \tab 7 \tab 106 \tab 17 \tab 1 \tab 8 \tab 2 \tab 29 \tab 1 \tab 2 \tab 20 \tab 9 \tab 2 \tab 1 \tab 4 \tab 5 \tab - \tab - \tab 1 \tab - \tab 2867 \cr
   2018 \tab 691 \tab - \tab 11 \tab 41 \tab 25 \tab 72 \tab 23 \tab 89 \tab 4 \tab 88 \tab 7 \tab 10 \tab 14 \tab 23 \tab 5 \tab 53 \tab 6 \tab 31 \tab 49 \tab 97 \tab 59 \tab - \tab - \tab 31 \tab 2 \tab 4 \tab 6 \tab 3 \tab 2 \tab 20 \tab 29 \tab 4 \tab 15 \tab 3 \tab 11 \tab 25 \tab 4 \tab 29 \tab 11 \tab 10 \tab 92 \tab 12 \tab 1 \tab 3 \tab 20 \tab 139 \tab 22 \tab 17 \tab 6 \tab 8 \tab 8 \tab 7 \tab 14 \tab 6 \tab 6 \tab 32 \tab 5 \tab 2 \tab 22 \tab 8 \tab 2 \tab - \tab 3 \tab 8 \tab 8 \tab 2 \tab 3 \tab 5 \tab 4 \tab 8 \tab 10 \tab 1 \tab 12 \tab 4 \tab 81 \tab 9 \tab 69 \tab 11 \tab 2 \tab 5 \tab 7 \tab 32 \tab 1 \tab - \tab 14 \tab 7 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 47277 \tab 71 \tab 119 \tab 609 \tab 188 \tab 1237 \tab 253 \tab 1155 \tab 31 \tab 1179 \tab 89 \tab 161 \tab 239 \tab 624 \tab 24 \tab 660 \tab 84 \tab 732 \tab 509 \tab 1027 \tab 656 \tab 27 \tab 2 \tab 214 \tab 40 \tab 32 \tab 95 \tab 23 \tab 47 \tab 192 \tab 420 \tab 154 \tab 157 \tab 14 \tab 174 \tab 279 \tab 75 \tab 503 \tab 118 \tab 110 \tab 1153 \tab 72 \tab 4 \tab 31 \tab 234 \tab 1806 \tab 190 \tab 216 \tab 62 \tab 54 \tab 65 \tab 70 \tab 138 \tab 62 \tab 105 \tab 202 \tab 80 \tab 7 \tab 267 \tab 170 \tab 33 \tab 22 \tab 18 \tab 145 \tab 48 \tab 24 \tab 33 \tab 65 \tab 30 \tab 90 \tab 74 \tab 9 \tab 94 \tab 58 \tab 1097 \tab 62 \tab 766 \tab 226 \tab 20 \tab 87 \tab 76 \tab 335 \tab 18 \tab 9 \tab 157 \tab 91 \tab 29 \tab 3 \tab 23 \tab 14 \tab 11 \tab 2 \tab 1 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
