% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{mapres80}
\alias{mapres80}
\title{Mothers occupational prestige score (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
mapres80
}
\details{
Question 10a. What kind of work did your (mother/MOTHER SUBSTITUTE) usually do while you were growing up? That is, what was her job called? b. What did she actually do?/What were some of her main duties?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab dk, na, iap \tab 17 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 64 \tab 65 \tab 66 \tab 68 \tab 69 \tab 73 \tab 74 \tab 75 \tab 86 \tab 63 \tab 70 \tab 71 \tab 72 \tab 67 \tab iap \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 1371 \tab 12 \tab 2 \tab 19 \tab 2 \tab 54 \tab 107 \tab 38 \tab 23 \tab 2 \tab 3 \tab 141 \tab 72 \tab 58 \tab 69 \tab 62 \tab 20 \tab 35 \tab 65 \tab 82 \tab 6 \tab 8 \tab 12 \tab 35 \tab 4 \tab 79 \tab 27 \tab 38 \tab 7 \tab 127 \tab 69 \tab 11 \tab 18 \tab 5 \tab 48 \tab 14 \tab 5 \tab 6 \tab 4 \tab 5 \tab 9 \tab 4 \tab 3 \tab 37 \tab 2 \tab 1 \tab 65 \tab 17 \tab 65 \tab 5 \tab 3 \tab 3 \tab 9 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 1176 \tab 14 \tab - \tab 25 \tab 5 \tab 41 \tab 102 \tab 45 \tab 17 \tab 2 \tab 1 \tab 119 \tab 68 \tab 33 \tab 62 \tab 59 \tab 37 \tab 47 \tab 63 \tab 89 \tab 3 \tab 9 \tab 21 \tab 35 \tab 11 \tab 84 \tab 28 \tab 27 \tab 9 \tab 151 \tab 87 \tab 10 \tab 19 \tab 5 \tab 82 \tab 18 \tab 1 \tab 5 \tab 8 \tab 8 \tab 9 \tab 2 \tab 9 \tab 34 \tab 4 \tab 2 \tab 80 \tab 19 \tab 96 \tab 6 \tab 6 \tab - \tab 6 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 1222 \tab 12 \tab - \tab 15 \tab 2 \tab 44 \tab 63 \tab 45 \tab 17 \tab 1 \tab 7 \tab 104 \tab 51 \tab 29 \tab 46 \tab 81 \tab 49 \tab 42 \tab 67 \tab 100 \tab 2 \tab 4 \tab 23 \tab 30 \tab 9 \tab 78 \tab 31 \tab 38 \tab 3 \tab 128 \tab 62 \tab 10 \tab 22 \tab 5 \tab 59 \tab 21 \tab 2 \tab 7 \tab 13 \tab 6 \tab 8 \tab 3 \tab 4 \tab 29 \tab 6 \tab 1 \tab 83 \tab 18 \tab 111 \tab - \tab 8 \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 1194 \tab 23 \tab 4 \tab 29 \tab 5 \tab 39 \tab 91 \tab 18 \tab 15 \tab 1 \tab 3 \tab 109 \tab 47 \tab 42 \tab 64 \tab 71 \tab 46 \tab 35 \tab 52 \tab 84 \tab 3 \tab 4 \tab 11 \tab 34 \tab 4 \tab 89 \tab 27 \tab 33 \tab 11 \tab 133 \tab 69 \tab 6 \tab 29 \tab - \tab 65 \tab 20 \tab - \tab 11 \tab 11 \tab 7 \tab 7 \tab 3 \tab 9 \tab 22 \tab 8 \tab - \tab 84 \tab 15 \tab 94 \tab 5 \tab 12 \tab 2 \tab 11 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 1146 \tab 14 \tab 5 \tab 22 \tab 9 \tab 39 \tab 83 \tab 14 \tab 13 \tab 3 \tab 3 \tab 125 \tab 38 \tab 28 \tab 53 \tab 87 \tab 57 \tab 48 \tab 63 \tab 77 \tab 1 \tab 4 \tab 16 \tab 30 \tab 5 \tab 77 \tab 21 \tab 39 \tab 13 \tab 160 \tab 68 \tab 9 \tab 30 \tab 2 \tab 60 \tab 19 \tab 6 \tab 10 \tab 11 \tab 4 \tab 7 \tab - \tab 8 \tab 16 \tab 7 \tab - \tab 74 \tab 14 \tab 100 \tab 1 \tab 6 \tab 3 \tab 9 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 2765 \cr
   2004 \tab - \tab 1116 \tab 21 \tab 1 \tab 30 \tab 9 \tab 40 \tab 82 \tab 26 \tab 13 \tab 1 \tab 7 \tab 111 \tab 39 \tab 30 \tab 59 \tab 91 \tab 52 \tab 41 \tab 58 \tab 61 \tab 4 \tab 7 \tab 23 \tab 26 \tab 6 \tab 78 \tab 44 \tab 36 \tab 15 \tab 122 \tab 52 \tab 8 \tab 35 \tab 8 \tab 79 \tab 15 \tab 2 \tab 9 \tab 12 \tab 5 \tab 12 \tab 6 \tab 4 \tab 18 \tab 10 \tab - \tab 105 \tab 31 \tab 114 \tab 9 \tab 4 \tab 3 \tab 10 \tab 2 \tab 4 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 2715 \tab 18 \tab 1 \tab 26 \tab 3 \tab 43 \tab 106 \tab 49 \tab 18 \tab 4 \tab 7 \tab 95 \tab 62 \tab 35 \tab 69 \tab 85 \tab 57 \tab 30 \tab 68 \tab 94 \tab 1 \tab 4 \tab 18 \tab 42 \tab 9 \tab 61 \tab 24 \tab 30 \tab 10 \tab 138 \tab 55 \tab 7 \tab 35 \tab 4 \tab 77 \tab 30 \tab 1 \tab 11 \tab 9 \tab 4 \tab 10 \tab - \tab 8 \tab 26 \tab 14 \tab 1 \tab 70 \tab 24 \tab 158 \tab 11 \tab 7 \tab 4 \tab 13 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 4510 \cr
   2008 \tab - \tab 745 \tab 15 \tab 1 \tab 24 \tab 1 \tab 30 \tab 59 \tab 19 \tab 20 \tab 1 \tab - \tab 64 \tab 48 \tab 20 \tab 54 \tab 67 \tab 52 \tab 39 \tab 36 \tab 61 \tab 5 \tab 7 \tab 11 \tab 27 \tab 9 \tab 48 \tab 29 \tab 17 \tab 12 \tab 102 \tab 58 \tab 5 \tab 19 \tab 5 \tab 65 \tab 10 \tab 2 \tab 3 \tab 9 \tab 2 \tab 8 \tab - \tab 6 \tab 15 \tab 7 \tab 4 \tab 51 \tab 15 \tab 92 \tab 2 \tab 4 \tab 7 \tab 4 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 701 \tab 19 \tab - \tab 20 \tab 5 \tab 38 \tab 51 \tab 7 \tab 14 \tab 4 \tab 4 \tab 95 \tab 39 \tab 28 \tab 43 \tab 55 \tab 59 \tab 33 \tab 35 \tab 65 \tab 1 \tab 4 \tab 9 \tab 29 \tab 8 \tab 56 \tab 30 \tab 23 \tab 8 \tab 92 \tab 57 \tab 5 \tab 27 \tab 3 \tab 81 \tab 18 \tab 1 \tab 12 \tab 4 \tab 4 \tab 9 \tab - \tab 5 \tab 15 \tab 6 \tab 3 \tab 66 \tab 21 \tab 102 \tab 5 \tab 3 \tab 2 \tab 8 \tab 1 \tab 7 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 21875 \tab 32658 \tab 148 \tab 14 \tab 210 \tab 41 \tab 368 \tab 744 \tab 261 \tab 150 \tab 19 \tab 35 \tab 963 \tab 464 \tab 303 \tab 519 \tab 658 \tab 429 \tab 350 \tab 507 \tab 713 \tab 26 \tab 51 \tab 144 \tab 288 \tab 65 \tab 650 \tab 261 \tab 281 \tab 88 \tab 1153 \tab 577 \tab 71 \tab 234 \tab 37 \tab 616 \tab 165 \tab 20 \tab 74 \tab 81 \tab 45 \tab 79 \tab 18 \tab 56 \tab 212 \tab 64 \tab 12 \tab 678 \tab 174 \tab 932 \tab 44 \tab 53 \tab 27 \tab 73 \tab 12 \tab 25 \tab 15 \tab 2 \tab 5 \tab 7 \tab 2 \tab 3544 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
