% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_588.R
\name{mar11}
\alias{mar11}
\title{Marital status of 11th person (visitor)}
\source{
General Social Survey https://gss.norc.org
}
\description{
mar11
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. d. IF 13 YEARS OR OLDER: Is \link{PERSON} now married, widowed, divorced, separated, or has he/she never been married?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab divorced \tab married \tab never married \tab separated \tab widowed \tab no answer \tab don't know \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1459 \tab 3 \tab 9 \tab 9 \tab 3 \tab 7 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1494 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1513 \tab 3 \tab 4 \tab 5 \tab - \tab 3 \tab 2 \tab - \tab - \tab 1530 \cr
   1978 \tab 1513 \tab 2 \tab 4 \tab 8 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1532 \cr
   1980 \tab 1441 \tab 5 \tab 5 \tab 7 \tab 5 \tab 4 \tab 1 \tab - \tab - \tab 1468 \cr
   1982 \tab 1845 \tab 2 \tab 3 \tab 4 \tab 3 \tab - \tab 3 \tab - \tab - \tab 1860 \cr
   1983 \tab 1576 \tab - \tab 2 \tab 9 \tab 2 \tab 7 \tab 3 \tab - \tab - \tab 1599 \cr
   1984 \tab 1443 \tab 3 \tab 7 \tab 9 \tab 4 \tab 2 \tab 5 \tab - \tab - \tab 1473 \cr
   1985 \tab 1501 \tab 3 \tab 6 \tab 5 \tab 2 \tab 4 \tab 13 \tab - \tab - \tab 1534 \cr
   1986 \tab 1454 \tab 1 \tab 3 \tab 9 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1470 \cr
   1987 \tab 1797 \tab 3 \tab 8 \tab 5 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 1819 \cr
   1988 \tab 1470 \tab 2 \tab 4 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1522 \tab 3 \tab 5 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1359 \tab - \tab 2 \tab 4 \tab 2 \tab 4 \tab 1 \tab - \tab - \tab 1372 \cr
   1991 \tab 1498 \tab 1 \tab 4 \tab 12 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1598 \tab 2 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2972 \tab 3 \tab 6 \tab 5 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 2992 \cr
   1996 \tab 2884 \tab - \tab 5 \tab 7 \tab 4 \tab - \tab 4 \tab - \tab - \tab 2904 \cr
   1998 \tab 2810 \tab 1 \tab 11 \tab 7 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2832 \cr
   2000 \tab 2810 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2817 \cr
   2002 \tab 2759 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2798 \tab 3 \tab 1 \tab 8 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2812 \cr
   2006 \tab 4463 \tab 4 \tab 11 \tab 28 \tab 1 \tab 3 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2007 \tab 4 \tab 2 \tab 8 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2019 \tab 2 \tab 6 \tab 14 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 1955 \tab 5 \tab 5 \tab 7 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2522 \tab 3 \tab 2 \tab 7 \tab 2 \tab - \tab - \tab 2 \tab - \tab 2538 \cr
   2016 \tab 2841 \tab 5 \tab 4 \tab 9 \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab 2867 \cr
   2018 \tab 2317 \tab 13 \tab 2 \tab 8 \tab 1 \tab 3 \tab 1 \tab 3 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3510 \tab 7 \tab 6 \tab 17 \tab 3 \tab 1 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 67751 \tab 86 \tab 132 \tab 227 \tab 49 \tab 61 \tab 45 \tab 7 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} married
\item \code{2} widowed
\item \code{3} divorced
\item \code{4} separated
\item \code{5} never married
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
