% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_588.R
\name{mar12}
\alias{mar12}
\title{mar12}
\source{
General Social Survey https://gss.norc.org
}
\description{
Marital status of 12th person (visitor)
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. d. IF 13 YEARS OR OLDER: Is \link{PERSON} now married, widowed, divorced, separated, or has he/she never been married?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab divorced \tab married \tab never married \tab separated \tab no answer \tab widowed \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1485 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1497 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1527 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1526 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1464 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 1468 \cr
   1982 \tab 1858 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1596 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1599 \cr
   1984 \tab 1466 \tab - \tab 5 \tab - \tab - \tab 2 \tab - \tab - \tab 1473 \cr
   1985 \tab 1530 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1534 \cr
   1986 \tab 1466 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1816 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1819 \cr
   1988 \tab 1480 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1532 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1369 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1513 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2988 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2901 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2821 \tab - \tab 6 \tab 4 \tab - \tab 1 \tab - \tab - \tab 2832 \cr
   2000 \tab 2816 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4499 \tab - \tab 7 \tab 3 \tab 1 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2021 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2036 \tab - \tab 3 \tab 3 \tab - \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 1971 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2863 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2343 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3542 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 68248 \tab 5 \tab 59 \tab 30 \tab 5 \tab 9 \tab 2 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} married
\item \code{2} widowed
\item \code{3} divorced
\item \code{4} separated
\item \code{5} never married
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
