% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_131.R
\name{marasian}
\alias{marasian}
\title{Close relative marry asian}
\source{
General Social Survey https://gss.norc.org
}
\description{
marasian
}
\details{
Question 405. What about having a close relative marry an Asian American person? Would you be in very favor of it happening, somewhat in favor, neither in favor nor opposed to it happening, somewhat opposed, or very opposed to it happening?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab don't know \tab favor \tab neither favor nor oppose \tab no answer \tab oppose \tab strongly favor \tab strongly oppose \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 24 \tab 90 \tab 632 \tab 9 \tab 373 \tab 39 \tab 205 \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1419 \tab 62 \tab 236 \tab 609 \tab 15 \tab 159 \tab 202 \tab 115 \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 3 \tab 155 \tab 482 \tab 3 \tab 99 \tab 106 \tab 58 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 11 \tab 335 \tab 1000 \tab 3 \tab 215 \tab 303 \tab 125 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 9 \tab 213 \tab 673 \tab 3 \tab 132 \tab 229 \tab 70 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 5 \tab 234 \tab 743 \tab 4 \tab 123 \tab 263 \tab 58 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 12 \tab 196 \tab 701 \tab 2 \tab 99 \tab 245 \tab 47 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 14 \tab 272 \tab 895 \tab 6 \tab 103 \tab 329 \tab 56 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 7 \tab 291 \tab 1021 \tab 5 \tab 118 \tab 396 \tab 50 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 7 \tab 258 \tab 793 \tab 6 \tab 87 \tab 367 \tab 41 \tab - \tab 2348 \cr
   2021 \tab 1318 \tab - \tab 242 \tab 1816 \tab 2 \tab 76 \tab 513 \tab 42 \tab 23 \tab 4032 \cr
   2022 \tab 1168 \tab 12 \tab 274 \tab 1354 \tab 4 \tab 79 \tab 604 \tab 48 \tab 1 \tab 3544 \cr
   Total \tab 52449 \tab 166 \tab 2796 \tab 10719 \tab 62 \tab 1663 \tab 3596 \tab 915 \tab 24 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} favor
\item \code{3} neither favor nor oppose
\item \code{4} oppose
\item \code{5} strongly oppose
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
