% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_601.R
\name{marcohrt}
\alias{marcohrt}
\title{marcohrt}
\source{
General Social Survey https://gss.norc.org
}
\description{
Year of first marriage
}
\details{
Question 1652. Marriage cohort of the respondent.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 1900 \tab 1904 \tab 1905 \tab 1907 \tab 1908 \tab 1909 \tab 1910 \tab 1911 \tab 1912 \tab 1913 \tab 1914 \tab 1915 \tab 1916 \tab 1917 \tab 1918 \tab 1919 \tab 1920 \tab 1921 \tab 1922 \tab 1923 \tab 1924 \tab 1925 \tab 1926 \tab 1927 \tab 1928 \tab 1929 \tab 1930 \tab 1931 \tab 1932 \tab 1933 \tab 1934 \tab 1935 \tab 1936 \tab 1937 \tab 1938 \tab 1939 \tab 1940 \tab 1941 \tab 1942 \tab 1943 \tab 1944 \tab 1945 \tab 1946 \tab 1947 \tab 1948 \tab 1949 \tab 1950 \tab 1951 \tab 1952 \tab 1953 \tab 1954 \tab 1955 \tab 1956 \tab 1957 \tab 1958 \tab 1959 \tab 1960 \tab 1961 \tab 1962 \tab 1963 \tab 1964 \tab 1965 \tab 1966 \tab 1967 \tab 1968 \tab 1969 \tab 1970 \tab 1971 \tab 1972 \tab iap \tab no answer \tab 1906 \tab 1901 \tab 1902 \tab 1973 \tab 1974 \tab 1903 \tab 1975 \tab 1976 \tab 1977 \tab 1978 \tab 1979 \tab 1980 \tab 1981 \tab 1982 \tab 1983 \tab 1984 \tab 1985 \tab 1986 \tab 1987 \tab 1988 \tab 1989 \tab 1990 \tab 1991 \tab 1992 \tab 1994 \tab 1993 \tab Total \cr
   1972 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 8 \tab 3 \tab 8 \tab 11 \tab 14 \tab 8 \tab 12 \tab 14 \tab 8 \tab 16 \tab 10 \tab 13 \tab 13 \tab 15 \tab 7 \tab 18 \tab 14 \tab 17 \tab 18 \tab 21 \tab 20 \tab 19 \tab 22 \tab 18 \tab 20 \tab 27 \tab 20 \tab 41 \tab 29 \tab 26 \tab 29 \tab 28 \tab 40 \tab 34 \tab 49 \tab 29 \tab 29 \tab 24 \tab 34 \tab 26 \tab 23 \tab 31 \tab 35 \tab 17 \tab 24 \tab 28 \tab 31 \tab 31 \tab 25 \tab 27 \tab 33 \tab 35 \tab 35 \tab 40 \tab 40 \tab 27 \tab 37 \tab 22 \tab 20 \tab 1 \tab 210 \tab 13 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 4 \tab 6 \tab 9 \tab 7 \tab 6 \tab 13 \tab 11 \tab 12 \tab 9 \tab 17 \tab 15 \tab 17 \tab 7 \tab 14 \tab 17 \tab 13 \tab 17 \tab 15 \tab 15 \tab 23 \tab 25 \tab 18 \tab 19 \tab 34 \tab 28 \tab 30 \tab 20 \tab 33 \tab 18 \tab 18 \tab 31 \tab 32 \tab 30 \tab 24 \tab 25 \tab 31 \tab 21 \tab 22 \tab 24 \tab 17 \tab 26 \tab 26 \tab 24 \tab 38 \tab 41 \tab 32 \tab 27 \tab 33 \tab 29 \tab 29 \tab 25 \tab 31 \tab 33 \tab 38 \tab 30 \tab 36 \tab 31 \tab 23 \tab 192 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 3 \tab 1 \tab 5 \tab 4 \tab 2 \tab 6 \tab 7 \tab 8 \tab 6 \tab 14 \tab 13 \tab 16 \tab 10 \tab 13 \tab 13 \tab 16 \tab 15 \tab 24 \tab 16 \tab 15 \tab 13 \tab 18 \tab 14 \tab 20 \tab 17 \tab 13 \tab 26 \tab 22 \tab 19 \tab 33 \tab 12 \tab 25 \tab 29 \tab 31 \tab 24 \tab 21 \tab 22 \tab 20 \tab 26 \tab 27 \tab 19 \tab 24 \tab 40 \tab 30 \tab 24 \tab 27 \tab 32 \tab 26 \tab 32 \tab 36 \tab 23 \tab 23 \tab 30 \tab 30 \tab 28 \tab 31 \tab 29 \tab 40 \tab 29 \tab 32 \tab 37 \tab 183 \tab 7 \tab 1 \tab 1 \tab 1 \tab 21 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab 4 \tab 4 \tab 5 \tab 4 \tab 5 \tab 5 \tab 4 \tab 10 \tab 9 \tab 7 \tab 9 \tab 15 \tab 1 \tab 14 \tab 25 \tab 14 \tab 23 \tab 13 \tab 16 \tab 10 \tab 16 \tab 18 \tab 18 \tab 17 \tab 20 \tab 22 \tab 23 \tab 25 \tab 22 \tab 19 \tab 11 \tab 24 \tab 24 \tab 35 \tab 21 \tab 32 \tab 21 \tab 33 \tab 19 \tab 19 \tab 20 \tab 28 \tab 33 \tab 29 \tab 24 \tab 28 \tab 27 \tab 23 \tab 26 \tab 21 \tab 19 \tab 33 \tab 28 \tab 31 \tab 23 \tab 39 \tab 36 \tab 36 \tab 40 \tab 37 \tab 211 \tab 14 \tab 1 \tab - \tab - \tab 23 \tab 14 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 3 \tab 6 \tab 4 \tab 6 \tab 7 \tab 7 \tab 5 \tab 8 \tab 6 \tab 11 \tab 12 \tab 7 \tab 12 \tab 12 \tab 10 \tab 13 \tab 7 \tab 10 \tab 14 \tab 14 \tab 16 \tab 15 \tab 22 \tab 17 \tab 20 \tab 24 \tab 27 \tab 15 \tab 22 \tab 30 \tab 20 \tab 24 \tab 19 \tab 14 \tab 26 \tab 29 \tab 23 \tab 18 \tab 24 \tab 25 \tab 17 \tab 21 \tab 20 \tab 20 \tab 25 \tab 18 \tab 28 \tab 19 \tab 27 \tab 28 \tab 23 \tab 32 \tab 21 \tab 35 \tab 30 \tab 24 \tab 23 \tab 37 \tab 39 \tab 38 \tab 30 \tab 37 \tab 205 \tab 8 \tab 1 \tab - \tab - \tab 33 \tab 32 \tab - \tab 17 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 4 \tab 7 \tab 7 \tab 6 \tab 6 \tab 13 \tab 13 \tab 13 \tab 13 \tab 10 \tab 12 \tab 9 \tab 9 \tab 15 \tab 13 \tab 16 \tab 5 \tab 17 \tab 16 \tab 14 \tab 15 \tab 17 \tab 22 \tab 32 \tab 28 \tab 26 \tab 21 \tab 19 \tab 21 \tab 37 \tab 31 \tab 28 \tab 20 \tab 37 \tab 29 \tab 26 \tab 24 \tab 19 \tab 32 \tab 18 \tab 25 \tab 30 \tab 22 \tab 23 \tab 32 \tab 25 \tab 27 \tab 39 \tab 28 \tab 28 \tab 29 \tab 32 \tab 28 \tab 32 \tab 34 \tab 31 \tab 225 \tab 11 \tab - \tab - \tab - \tab 30 \tab 29 \tab - \tab 19 \tab 22 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 3 \tab 4 \tab 4 \tab 2 \tab 6 \tab 4 \tab 4 \tab 10 \tab 2 \tab 12 \tab 5 \tab 11 \tab 11 \tab 15 \tab 13 \tab 18 \tab 14 \tab 12 \tab 10 \tab 17 \tab 8 \tab 15 \tab 14 \tab 21 \tab 9 \tab 13 \tab 16 \tab 17 \tab 22 \tab 29 \tab 21 \tab 19 \tab 26 \tab 34 \tab 23 \tab 26 \tab 12 \tab 28 \tab 12 \tab 17 \tab 21 \tab 19 \tab 24 \tab 31 \tab 21 \tab 16 \tab 16 \tab 33 \tab 18 \tab 35 \tab 20 \tab 40 \tab 32 \tab 29 \tab 36 \tab 34 \tab 37 \tab 41 \tab 24 \tab 35 \tab 237 \tab 12 \tab - \tab - \tab - \tab 45 \tab 38 \tab - \tab 33 \tab 26 \tab 16 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab 5 \tab 3 \tab 7 \tab 5 \tab 1 \tab 3 \tab 5 \tab 4 \tab 4 \tab 4 \tab 14 \tab 9 \tab 10 \tab 14 \tab 13 \tab 14 \tab 8 \tab 11 \tab 17 \tab 19 \tab 17 \tab 21 \tab 16 \tab 15 \tab 14 \tab 15 \tab 26 \tab 11 \tab 13 \tab 20 \tab 29 \tab 29 \tab 18 \tab 22 \tab 19 \tab 28 \tab 15 \tab 16 \tab 28 \tab 23 \tab 26 \tab 21 \tab 18 \tab 19 \tab 20 \tab 19 \tab 13 \tab 21 \tab 32 \tab 31 \tab 26 \tab 18 \tab 29 \tab 32 \tab 29 \tab 31 \tab 30 \tab 31 \tab 240 \tab 20 \tab - \tab - \tab - \tab 29 \tab 24 \tab - \tab 33 \tab 30 \tab 33 \tab 23 \tab 16 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 3 \tab 7 \tab 4 \tab 2 \tab 8 \tab 5 \tab 12 \tab 3 \tab 12 \tab 4 \tab 13 \tab 7 \tab 10 \tab 18 \tab 15 \tab 11 \tab 12 \tab 16 \tab 15 \tab 14 \tab 17 \tab 20 \tab 21 \tab 25 \tab 30 \tab 19 \tab 30 \tab 28 \tab 22 \tab 20 \tab 44 \tab 26 \tab 29 \tab 30 \tab 27 \tab 22 \tab 19 \tab 31 \tab 20 \tab 16 \tab 30 \tab 23 \tab 17 \tab 25 \tab 28 \tab 24 \tab 21 \tab 14 \tab 37 \tab 28 \tab 35 \tab 33 \tab 28 \tab 26 \tab 37 \tab 35 \tab 42 \tab 348 \tab 24 \tab - \tab - \tab - \tab 33 \tab 29 \tab - \tab 35 \tab 34 \tab 41 \tab 31 \tab 48 \tab 34 \tab 21 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 4 \tab 2 \tab 2 \tab 2 \tab 6 \tab 2 \tab 6 \tab 7 \tab 4 \tab 9 \tab 11 \tab 12 \tab 8 \tab 6 \tab 6 \tab 12 \tab 13 \tab 16 \tab 22 \tab 18 \tab 9 \tab 14 \tab 21 \tab 10 \tab 27 \tab 14 \tab 11 \tab 20 \tab 36 \tab 35 \tab 20 \tab 20 \tab 25 \tab 20 \tab 21 \tab 20 \tab 19 \tab 21 \tab 22 \tab 22 \tab 23 \tab 21 \tab 24 \tab 21 \tab 21 \tab 18 \tab 24 \tab 23 \tab 30 \tab 25 \tab 35 \tab 31 \tab 35 \tab 28 \tab 34 \tab 29 \tab 264 \tab 17 \tab - \tab - \tab - \tab 39 \tab 35 \tab - \tab 35 \tab 31 \tab 38 \tab 37 \tab 23 \tab 35 \tab 23 \tab 20 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab 2 \tab 4 \tab 7 \tab 3 \tab 4 \tab 8 \tab 4 \tab 9 \tab 11 \tab 8 \tab 5 \tab 13 \tab 7 \tab 11 \tab 8 \tab 14 \tab 11 \tab 15 \tab 17 \tab 10 \tab 20 \tab 17 \tab 13 \tab 19 \tab 18 \tab 22 \tab 24 \tab 15 \tab 15 \tab 14 \tab 18 \tab 17 \tab 18 \tab 20 \tab 27 \tab 15 \tab 17 \tab 24 \tab 20 \tab 17 \tab 15 \tab 20 \tab 15 \tab 21 \tab 23 \tab 30 \tab 27 \tab 22 \tab 22 \tab 28 \tab 36 \tab 34 \tab 32 \tab 282 \tab 21 \tab - \tab - \tab - \tab 41 \tab 16 \tab - \tab 22 \tab 23 \tab 28 \tab 24 \tab 38 \tab 17 \tab 23 \tab 32 \tab 28 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab 6 \tab 5 \tab 7 \tab 9 \tab 4 \tab 9 \tab 12 \tab 9 \tab 11 \tab 5 \tab 11 \tab 13 \tab 9 \tab 13 \tab 12 \tab 12 \tab 13 \tab 21 \tab 15 \tab 15 \tab 14 \tab 18 \tab 17 \tab 17 \tab 21 \tab 21 \tab 29 \tab 32 \tab 20 \tab 21 \tab 20 \tab 18 \tab 14 \tab 28 \tab 20 \tab 20 \tab 13 \tab 22 \tab 22 \tab 25 \tab 14 \tab 28 \tab 11 \tab 28 \tab 35 \tab 28 \tab 24 \tab 18 \tab 27 \tab 26 \tab 35 \tab 25 \tab 266 \tab 13 \tab - \tab - \tab - \tab 26 \tab 36 \tab - \tab 24 \tab 29 \tab 21 \tab 33 \tab 21 \tab 27 \tab 23 \tab 29 \tab 35 \tab 19 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 5 \tab 3 \tab 2 \tab 3 \tab 4 \tab 9 \tab 3 \tab 11 \tab 8 \tab 8 \tab 8 \tab 5 \tab 7 \tab 12 \tab 14 \tab 15 \tab 11 \tab 11 \tab 8 \tab 20 \tab 25 \tab 15 \tab 22 \tab 16 \tab 20 \tab 18 \tab 27 \tab 18 \tab 21 \tab 18 \tab 18 \tab 15 \tab 18 \tab 14 \tab 21 \tab 16 \tab 13 \tab 16 \tab 22 \tab 21 \tab 10 \tab 19 \tab 21 \tab 18 \tab 20 \tab 21 \tab 31 \tab 26 \tab 25 \tab 39 \tab 32 \tab 27 \tab 29 \tab 263 \tab 14 \tab - \tab - \tab - \tab 31 \tab 29 \tab - \tab 25 \tab 24 \tab 29 \tab 24 \tab 22 \tab 35 \tab 39 \tab 22 \tab 16 \tab 17 \tab 14 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 2 \tab 2 \tab 4 \tab 3 \tab 8 \tab 5 \tab 2 \tab 3 \tab 5 \tab 10 \tab 9 \tab 7 \tab 10 \tab 11 \tab 11 \tab 18 \tab 17 \tab 13 \tab 15 \tab 17 \tab 20 \tab 19 \tab 14 \tab 19 \tab 16 \tab 18 \tab 29 \tab 23 \tab 24 \tab 24 \tab 17 \tab 23 \tab 27 \tab 16 \tab 23 \tab 18 \tab 13 \tab 20 \tab 31 \tab 20 \tab 28 \tab 24 \tab 22 \tab 20 \tab 30 \tab 34 \tab 24 \tab 32 \tab 33 \tab 40 \tab 27 \tab 42 \tab 26 \tab 37 \tab 380 \tab 27 \tab - \tab - \tab - \tab 24 \tab 31 \tab - \tab 30 \tab 27 \tab 42 \tab 21 \tab 24 \tab 24 \tab 27 \tab 32 \tab 28 \tab 32 \tab 38 \tab 19 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 4 \tab 4 \tab 4 \tab 6 \tab 4 \tab 5 \tab 7 \tab 7 \tab 5 \tab 5 \tab 7 \tab 11 \tab 9 \tab 8 \tab 14 \tab 11 \tab 11 \tab 13 \tab 39 \tab 11 \tab 15 \tab 23 \tab 15 \tab 24 \tab 15 \tab 34 \tab 17 \tab 19 \tab 13 \tab 22 \tab 13 \tab 15 \tab 18 \tab 11 \tab 11 \tab 22 \tab 8 \tab 14 \tab 13 \tab 21 \tab 15 \tab 20 \tab 21 \tab 22 \tab 26 \tab 20 \tab 24 \tab 24 \tab 31 \tab 37 \tab 33 \tab 297 \tab 16 \tab - \tab - \tab - \tab 31 \tab 25 \tab - \tab 26 \tab 23 \tab 21 \tab 23 \tab 31 \tab 27 \tab 24 \tab 19 \tab 18 \tab 17 \tab 33 \tab 17 \tab 21 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 3 \tab 5 \tab 3 \tab 2 \tab 6 \tab 9 \tab 3 \tab 8 \tab 10 \tab 5 \tab 11 \tab 6 \tab 8 \tab 16 \tab 10 \tab 12 \tab 10 \tab 21 \tab 17 \tab 22 \tab 15 \tab 13 \tab 25 \tab 29 \tab 22 \tab 15 \tab 21 \tab 19 \tab 18 \tab 13 \tab 13 \tab 8 \tab 24 \tab 20 \tab 21 \tab 17 \tab 25 \tab 22 \tab 28 \tab 14 \tab 22 \tab 30 \tab 25 \tab 21 \tab 22 \tab 20 \tab 21 \tab 31 \tab 28 \tab 31 \tab 301 \tab 17 \tab - \tab - \tab - \tab 20 \tab 34 \tab - \tab 22 \tab 26 \tab 22 \tab 23 \tab 30 \tab 21 \tab 24 \tab 28 \tab 33 \tab 19 \tab 23 \tab 20 \tab 28 \tab 18 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 4 \tab 1 \tab 9 \tab 2 \tab 11 \tab 3 \tab 5 \tab 4 \tab 5 \tab 4 \tab 11 \tab 9 \tab 13 \tab 18 \tab 10 \tab 12 \tab 11 \tab 10 \tab 13 \tab 11 \tab 13 \tab 19 \tab 24 \tab 22 \tab 18 \tab 14 \tab 19 \tab 8 \tab 14 \tab 9 \tab 13 \tab 12 \tab 8 \tab 16 \tab 15 \tab 19 \tab 15 \tab 14 \tab 16 \tab 19 \tab 25 \tab 21 \tab 24 \tab 23 \tab 18 \tab 19 \tab 34 \tab 24 \tab 24 \tab 265 \tab 18 \tab - \tab - \tab - \tab 28 \tab 33 \tab - \tab 30 \tab 22 \tab 20 \tab 28 \tab 29 \tab 21 \tab 23 \tab 18 \tab 21 \tab 18 \tab 22 \tab 21 \tab 20 \tab 24 \tab 20 \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 4 \tab 3 \tab 7 \tab 5 \tab 6 \tab 9 \tab 4 \tab 10 \tab 8 \tab 2 \tab 6 \tab 14 \tab 8 \tab 7 \tab 16 \tab 18 \tab 19 \tab 15 \tab 9 \tab 14 \tab 16 \tab 20 \tab 16 \tab 25 \tab 16 \tab 10 \tab 14 \tab 18 \tab 24 \tab 15 \tab 19 \tab 15 \tab 13 \tab 12 \tab 20 \tab 13 \tab 14 \tab 17 \tab 19 \tab 20 \tab 20 \tab 25 \tab 24 \tab 17 \tab 28 \tab 22 \tab 37 \tab 31 \tab 324 \tab 16 \tab - \tab - \tab - \tab 19 \tab 28 \tab - \tab 29 \tab 23 \tab 32 \tab 22 \tab 25 \tab 23 \tab 32 \tab 26 \tab 28 \tab 26 \tab 24 \tab 25 \tab 16 \tab 18 \tab 26 \tab 20 \tab 2 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 4 \tab 5 \tab 2 \tab 7 \tab - \tab 3 \tab 3 \tab 10 \tab 5 \tab 13 \tab 14 \tab 12 \tab 11 \tab 18 \tab 13 \tab 13 \tab 21 \tab 9 \tab 17 \tab 24 \tab 21 \tab 20 \tab 19 \tab 16 \tab 17 \tab 20 \tab 17 \tab 16 \tab 14 \tab 21 \tab 16 \tab 13 \tab 14 \tab 18 \tab 18 \tab 18 \tab 18 \tab 23 \tab 22 \tab 20 \tab 20 \tab 33 \tab 22 \tab 31 \tab 41 \tab 27 \tab 299 \tab 16 \tab - \tab - \tab - \tab 26 \tab 27 \tab - \tab 27 \tab 27 \tab 29 \tab 25 \tab 28 \tab 33 \tab 24 \tab 29 \tab 27 \tab 28 \tab 28 \tab 32 \tab 27 \tab 30 \tab 26 \tab 27 \tab 26 \tab 18 \tab 2 \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 6 \tab 4 \tab 4 \tab 5 \tab 13 \tab 12 \tab 14 \tab 14 \tab 16 \tab 15 \tab 10 \tab 8 \tab 15 \tab 24 \tab 13 \tab 11 \tab 17 \tab 17 \tab 16 \tab 19 \tab 12 \tab 20 \tab 14 \tab 18 \tab 12 \tab 20 \tab 18 \tab 16 \tab 17 \tab 17 \tab 22 \tab 23 \tab 27 \tab 21 \tab 18 \tab 20 \tab 25 \tab 36 \tab 24 \tab 21 \tab 1786 \tab 19 \tab - \tab - \tab - \tab 19 \tab 20 \tab - \tab 25 \tab 30 \tab 31 \tab 27 \tab 27 \tab 32 \tab 21 \tab 31 \tab 35 \tab 25 \tab 24 \tab 28 \tab 21 \tab 26 \tab 21 \tab 27 \tab 16 \tab 20 \tab - \tab 11 \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 1 \tab 3 \tab 3 \tab 4 \tab 7 \tab 14 \tab 11 \tab 14 \tab 18 \tab 26 \tab 31 \tab 46 \tab 46 \tab 74 \tab 69 \tab 63 \tab 100 \tab 102 \tab 126 \tab 124 \tab 151 \tab 137 \tab 168 \tab 178 \tab 211 \tab 213 \tab 196 \tab 193 \tab 210 \tab 236 \tab 268 \tab 286 \tab 307 \tab 296 \tab 331 \tab 354 \tab 449 \tab 392 \tab 389 \tab 353 \tab 358 \tab 490 \tab 564 \tab 476 \tab 448 \tab 400 \tab 453 \tab 391 \tab 372 \tab 395 \tab 422 \tab 427 \tab 399 \tab 421 \tab 417 \tab 457 \tab 441 \tab 432 \tab 437 \tab 442 \tab 579 \tab 539 \tab 544 \tab 544 \tab 564 \tab 597 \tab 651 \tab 623 \tab 593 \tab 46788 \tab 307 \tab 4 \tab 1 \tab 1 \tab 518 \tab 481 \tab 1 \tab 432 \tab 398 \tab 403 \tab 342 \tab 362 \tab 330 \tab 304 \tab 288 \tab 270 \tab 202 \tab 207 \tab 162 \tab 134 \tab 117 \tab 93 \tab 74 \tab 44 \tab 38 \tab 2 \tab 11 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
