% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_03.R
\name{marital}
\alias{marital}
\title{Marital status}
\source{
General Social Survey https://gss.norc.org
}
\description{
marital
}
\details{
Question 4. Are you currently -- married, widowed, divorced, separated, or have you never been married?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab divorced \tab married \tab never married \tab separated \tab widowed \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 65 \tab 1160 \tab 210 \tab 39 \tab 139 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 60 \tab 1076 \tab 193 \tab 52 \tab 123 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 76 \tab 1065 \tab 184 \tab 38 \tab 121 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 84 \tab 1002 \tab 211 \tab 49 \tab 144 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 98 \tab 974 \tab 207 \tab 51 \tab 169 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 103 \tab 975 \tab 225 \tab 62 \tab 165 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 134 \tab 960 \tab 237 \tab 47 \tab 153 \tab 1 \tab - \tab - \tab 1532 \cr
   1980 \tab 125 \tab 889 \tab 242 \tab 48 \tab 164 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 203 \tab 993 \tab 351 \tab 91 \tab 222 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 159 \tab 965 \tab 265 \tab 48 \tab 162 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 166 \tab 829 \tab 282 \tab 42 \tab 154 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 169 \tab 871 \tab 268 \tab 65 \tab 161 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 148 \tab 826 \tab 266 \tab 60 \tab 170 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 222 \tab 915 \tab 380 \tab 84 \tab 217 \tab 1 \tab - \tab - \tab 1819 \cr
   1988 \tab 186 \tab 788 \tab 298 \tab 48 \tab 161 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 177 \tab 847 \tab 301 \tab 59 \tab 153 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 171 \tab 727 \tab 265 \tab 37 \tab 171 \tab 1 \tab - \tab - \tab 1372 \cr
   1991 \tab 175 \tab 804 \tab 325 \tab 45 \tab 168 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 230 \tab 859 \tab 300 \tab 44 \tab 172 \tab 1 \tab - \tab - \tab 1606 \cr
   1994 \tab 446 \tab 1541 \tab 614 \tab 102 \tab 288 \tab 1 \tab - \tab - \tab 2992 \cr
   1996 \tab 455 \tab 1390 \tab 658 \tab 118 \tab 282 \tab 1 \tab - \tab - \tab 2904 \cr
   1998 \tab 446 \tab 1346 \tab 663 \tab 93 \tab 283 \tab 1 \tab - \tab - \tab 2832 \cr
   2000 \tab 441 \tab 1278 \tab 712 \tab 112 \tab 273 \tab 1 \tab - \tab - \tab 2817 \cr
   2002 \tab 445 \tab 1269 \tab 708 \tab 96 \tab 247 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 415 \tab 1479 \tab 619 \tab 95 \tab 204 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 732 \tab 2170 \tab 1080 \tab 156 \tab 366 \tab 6 \tab - \tab - \tab 4510 \cr
   2008 \tab 281 \tab 972 \tab 531 \tab 70 \tab 164 \tab 5 \tab - \tab - \tab 2023 \cr
   2010 \tab 341 \tab 891 \tab 565 \tab 65 \tab 181 \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab 317 \tab 900 \tab 526 \tab 68 \tab 163 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 411 \tab 1158 \tab 675 \tab 81 \tab 209 \tab 4 \tab - \tab - \tab 2538 \cr
   2016 \tab 495 \tab 1212 \tab 806 \tab 102 \tab 251 \tab 1 \tab - \tab - \tab 2867 \cr
   2018 \tab 403 \tab 998 \tab 670 \tab 75 \tab 200 \tab 2 \tab - \tab - \tab 2348 \cr
   2021 \tab 655 \tab 1999 \tab 972 \tab 96 \tab 301 \tab 1 \tab 1 \tab 7 \tab 4032 \cr
   2022 \tab 608 \tab 1462 \tab 1101 \tab 103 \tab 255 \tab 9 \tab 1 \tab 5 \tab 3544 \cr
   Total \tab 9642 \tab 37590 \tab 15910 \tab 2441 \tab 6756 \tab 37 \tab 2 \tab 12 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} married
\item \code{2} widowed
\item \code{3} divorced
\item \code{4} separated
\item \code{5} never married
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
