% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_598.R
\name{masei10}
\alias{masei10}
\title{R's mother's socioeconomic index (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
masei10
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 10.6 \tab 11.8 \tab 12.4 \tab 12.6 \tab 12.7 \tab 13.2 \tab 13.3 \tab 13.6 \tab 14 \tab 14.6 \tab 14.8 \tab 15.8 \tab 17 \tab 17.1 \tab 17.5 \tab 18.6 \tab 18.8 \tab 19.6 \tab 19.7 \tab 20 \tab 20.1 \tab 20.3 \tab 20.5 \tab 20.7 \tab 20.8 \tab 21.1 \tab 21.6 \tab 21.8 \tab 22.1 \tab 22.4 \tab 22.9 \tab 23.3 \tab 23.8 \tab 24.1 \tab 24.2 \tab 24.6 \tab 25 \tab 25.1 \tab 25.2 \tab 25.7 \tab 26.1 \tab 26.2 \tab 26.5 \tab 26.6 \tab 26.8 \tab 26.9 \tab 27 \tab 28.4 \tab 28.5 \tab 28.6 \tab 28.7 \tab 29.2 \tab 29.4 \tab 29.9 \tab 30.9 \tab 31.1 \tab 31.6 \tab 31.9 \tab 32 \tab 32.6 \tab 32.7 \tab 33.2 \tab 33.4 \tab 34.4 \tab 34.5 \tab 34.6 \tab 34.7 \tab 35.1 \tab 35.2 \tab 35.3 \tab 35.4 \tab 35.8 \tab 36.4 \tab 36.5 \tab 36.8 \tab 36.9 \tab 37.2 \tab 37.6 \tab 37.7 \tab 37.9 \tab 38.1 \tab 38.2 \tab 38.3 \tab 38.8 \tab 38.9 \tab 39.2 \tab 39.5 \tab 39.6 \tab 39.7 \tab 39.9 \tab 41 \tab 41.1 \tab 41.8 \tab 42.9 \tab 43 \tab 44.6 \tab 44.9 \tab 45 \tab 45.4 \tab 46 \tab 46.5 \tab 46.6 \tab 47.9 \tab 48.3 \tab 48.4 \tab 48.8 \tab 50 \tab 50.1 \tab 50.3 \tab 50.4 \tab 51.7 \tab 51.8 \tab 52 \tab 52.1 \tab 52.4 \tab 53 \tab 53.1 \tab 53.8 \tab 53.9 \tab 54.6 \tab 55.3 \tab 56.8 \tab 56.9 \tab 57 \tab 57.4 \tab 57.5 \tab 57.8 \tab 58.2 \tab 58.4 \tab 59 \tab 59.1 \tab 59.8 \tab 59.9 \tab 60.1 \tab 60.2 \tab 60.3 \tab 60.5 \tab 61.2 \tab 61.4 \tab 62 \tab 62.1 \tab 62.6 \tab 62.9 \tab 63.3 \tab 63.7 \tab 64.5 \tab 65.3 \tab 65.6 \tab 66.2 \tab 67.7 \tab 68.6 \tab 69.3 \tab 70.4 \tab 71 \tab 71.7 \tab 71.8 \tab 72.5 \tab 73.4 \tab 73.9 \tab 75.5 \tab 75.7 \tab 76.3 \tab 76.7 \tab 77 \tab 77.2 \tab 77.3 \tab 77.4 \tab 77.9 \tab 80.7 \tab 80.9 \tab 82.5 \tab 84.2 \tab 84.3 \tab 84.5 \tab 85.8 \tab 86.6 \tab 86.8 \tab 89.9 \tab 9 \tab 91.1 \tab 92 \tab 92.8 \tab 20.9 \tab 21 \tab 21.5 \tab 22.5 \tab 23.7 \tab 24 \tab 29.5 \tab 29.6 \tab 29.7 \tab 30.3 \tab 31.2 \tab 31.3 \tab 34.8 \tab 35.5 \tab 37.1 \tab 43.1 \tab 49.2 \tab 50.7 \tab 52.3 \tab 55.1 \tab 59.4 \tab 61.5 \tab 62.4 \tab 62.7 \tab 64.2 \tab 67.8 \tab 69.4 \tab 70.5 \tab 72.7 \tab 73.6 \tab 79.9 \tab 81 \tab 81.9 \tab 83.8 \tab 85 \tab 86.5 \tab 88.3 \tab 88.6 \tab 88.7 \tab 89.4 \tab 92.2 \tab 19.2 \tab 23.2 \tab 25.6 \tab 27.1 \tab 27.4 \tab 27.5 \tab 27.6 \tab 29.3 \tab 41.4 \tab 42.8 \tab 43.8 \tab 45.1 \tab 45.3 \tab 49.3 \tab 49.4 \tab 52.6 \tab 53.6 \tab 53.7 \tab 58.6 \tab 63.1 \tab 63.2 \tab 64.4 \tab 64.9 \tab 68 \tab 74.8 \tab 18.9 \tab 19 \tab 24.8 \tab 25.4 \tab 27.7 \tab 37.8 \tab 42.7 \tab 46.2 \tab 55 \tab 61.9 \tab 70.3 \tab 71.6 \tab 73 \tab 73.2 \tab 75.6 \tab 80.2 \tab 82.4 \tab 84 \tab 85.2 \tab 86.1 \tab 87.3 \tab 87.6 \tab 87.9 \tab 88 \tab 92.1 \tab 21.3 \tab 31.5 \tab 33.9 \tab 36.7 \tab 40.1 \tab 48.1 \tab 51.5 \tab 52.5 \tab 73.5 \tab 81.5 \tab 82.7 \tab 91.6 \tab 91.9 \tab 30 \tab 33.7 \tab 44.5 \tab 52.9 \tab 53.2 \tab 54.1 \tab 56.3 \tab 57.1 \tab 66.4 \tab 70.1 \tab 74.6 \tab 77.7 \tab 78.3 \tab 87.8 \tab 30.8 \tab 39.1 \tab 54.3 \tab 37.4 \tab 66.1 \tab 69.5 \tab 71.4 \tab 73.1 \tab 83.4 \tab 86.7 \tab 88.4 \tab 57.3 \tab 65.1 \tab 85.5 \tab 86.9 \tab 34.1 \tab 36.2 \tab 37 \tab 49.6 \tab 72.6 \tab 81.6 \tab 23 \tab 43.5 \tab 79.6 \tab 28.8 \tab 40.4 \tab 58.1 \tab 27.2 \tab 33.3 \tab 41.2 \tab 79.3 \tab 85.9 \tab 37.3 \tab 76.5 \tab 89.3 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1351 \tab 1 \tab 15 \tab 7 \tab 106 \tab 7 \tab 66 \tab 5 \tab 51 \tab 26 \tab 3 \tab 6 \tab 14 \tab 1 \tab 12 \tab 1 \tab 3 \tab 4 \tab 51 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 19 \tab 7 \tab 1 \tab 41 \tab 21 \tab 2 \tab 1 \tab 3 \tab 27 \tab 11 \tab 2 \tab 73 \tab 26 \tab 31 \tab 3 \tab 1 \tab 10 \tab 17 \tab 1 \tab 21 \tab 1 \tab 43 \tab 7 \tab 10 \tab 3 \tab 2 \tab 8 \tab 1 \tab 1 \tab 2 \tab 5 \tab 2 \tab 2 \tab 8 \tab 9 \tab 3 \tab 16 \tab 3 \tab 2 \tab 3 \tab 1 \tab 2 \tab 2 \tab 3 \tab 6 \tab 1 \tab 1 \tab 1 \tab 18 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab 18 \tab 3 \tab 53 \tab 138 \tab 1 \tab 3 \tab 1 \tab 2 \tab 75 \tab 11 \tab 13 \tab 7 \tab 3 \tab 3 \tab 36 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 12 \tab 23 \tab 5 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 12 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 9 \tab 2 \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 4 \tab 3 \tab 1 \tab 9 \tab 2 \tab 1 \tab 5 \tab 1 \tab 10 \tab 3 \tab 7 \tab 2 \tab 3 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 16 \tab 6 \tab 4 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 15 \tab 1 \tab 2 \tab 4 \tab 1 \tab 3 \tab 1 \tab 2 \tab 62 \tab 18 \tab 52 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1158 \tab 1 \tab 4 \tab 11 \tab 105 \tab 9 \tab 71 \tab 14 \tab 48 \tab 23 \tab 2 \tab 6 \tab 21 \tab 1 \tab 15 \tab 1 \tab 2 \tab 4 \tab 50 \tab 4 \tab - \tab 3 \tab - \tab 3 \tab 16 \tab 13 \tab 4 \tab 42 \tab 20 \tab - \tab - \tab 3 \tab 19 \tab 10 \tab 2 \tab 79 \tab 26 \tab 29 \tab 2 \tab - \tab 11 \tab 10 \tab 3 \tab 19 \tab 1 \tab 52 \tab 3 \tab 20 \tab 5 \tab 1 \tab 7 \tab 3 \tab - \tab - \tab 9 \tab 2 \tab 4 \tab 9 \tab 8 \tab 5 \tab 20 \tab - \tab 5 \tab 5 \tab - \tab 1 \tab - \tab 2 \tab 6 \tab - \tab 5 \tab - \tab 15 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 6 \tab 4 \tab - \tab 14 \tab 3 \tab 57 \tab 152 \tab 3 \tab 1 \tab - \tab 2 \tab 65 \tab 13 \tab 14 \tab 8 \tab 4 \tab 3 \tab 28 \tab 1 \tab - \tab - \tab - \tab - \tab 12 \tab 22 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab 3 \tab - \tab - \tab - \tab 18 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 2 \tab 14 \tab 1 \tab 2 \tab 4 \tab 2 \tab - \tab 5 \tab - \tab 6 \tab 9 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 2 \tab 13 \tab 4 \tab 6 \tab - \tab 10 \tab - \tab 9 \tab - \tab 3 \tab - \tab 1 \tab 9 \tab 3 \tab 6 \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 18 \tab 1 \tab - \tab 4 \tab - \tab 7 \tab - \tab 1 \tab 70 \tab 26 \tab 57 \tab - \tab 8 \tab - \tab 1 \tab - \tab - \tab 5 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1213 \tab - \tab 13 \tab 7 \tab 87 \tab 4 \tab 67 \tab 5 \tab 44 \tab 24 \tab 2 \tab 16 \tab 6 \tab 1 \tab 14 \tab - \tab - \tab 8 \tab 42 \tab 5 \tab - \tab 5 \tab - \tab 2 \tab 13 \tab 12 \tab 3 \tab 43 \tab 27 \tab - \tab - \tab 1 \tab 20 \tab 9 \tab - \tab 74 \tab 21 \tab 30 \tab 2 \tab - \tab 15 \tab 18 \tab 1 \tab 19 \tab 1 \tab 62 \tab 2 \tab 9 \tab 2 \tab 1 \tab 4 \tab 3 \tab - \tab - \tab 7 \tab 2 \tab 2 \tab 9 \tab 6 \tab 2 \tab 12 \tab - \tab 2 \tab 3 \tab 2 \tab 3 \tab 3 \tab 2 \tab 5 \tab - \tab 2 \tab - \tab 23 \tab - \tab - \tab - \tab 1 \tab 5 \tab 10 \tab - \tab 4 \tab 10 \tab 2 \tab 40 \tab 125 \tab - \tab 3 \tab 1 \tab 4 \tab 60 \tab 10 \tab 18 \tab 13 \tab 6 \tab 1 \tab 32 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 8 \tab 18 \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 12 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab 10 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 7 \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab 2 \tab 12 \tab 2 \tab 5 \tab 1 \tab 15 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 9 \tab - \tab 10 \tab - \tab - \tab 6 \tab 4 \tab - \tab 2 \tab 1 \tab 7 \tab - \tab 13 \tab - \tab - \tab 6 \tab 1 \tab 4 \tab - \tab - \tab 74 \tab 26 \tab 62 \tab - \tab 5 \tab 3 \tab 2 \tab 1 \tab - \tab 10 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 9 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1184 \tab 5 \tab 5 \tab 6 \tab 102 \tab 3 \tab 56 \tab 5 \tab 54 \tab 23 \tab 3 \tab 15 \tab 9 \tab 1 \tab 11 \tab 2 \tab 1 \tab 3 \tab 40 \tab 3 \tab 1 \tab 6 \tab - \tab 5 \tab 13 \tab 10 \tab 2 \tab 39 \tab 28 \tab - \tab 1 \tab 1 \tab 11 \tab 5 \tab - \tab 71 \tab 21 \tab 29 \tab 1 \tab - \tab 5 \tab 12 \tab - \tab 30 \tab 1 \tab 50 \tab 7 \tab 10 \tab 5 \tab 2 \tab 9 \tab 1 \tab - \tab - \tab 6 \tab 5 \tab 2 \tab 11 \tab 3 \tab 3 \tab 17 \tab 2 \tab 3 \tab 3 \tab 1 \tab - \tab 5 \tab 6 \tab 2 \tab - \tab 3 \tab - \tab 18 \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 10 \tab 2 \tab 2 \tab 8 \tab - \tab 44 \tab 151 \tab - \tab 1 \tab - \tab 2 \tab 63 \tab 12 \tab 17 \tab 11 \tab 5 \tab 2 \tab 35 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 9 \tab 19 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 12 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 14 \tab - \tab 1 \tab 4 \tab 5 \tab 2 \tab 3 \tab - \tab 12 \tab 9 \tab 3 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 12 \tab 3 \tab 1 \tab - \tab 11 \tab 2 \tab 7 \tab - \tab 6 \tab - \tab 1 \tab 12 \tab 1 \tab 9 \tab - \tab - \tab 4 \tab 4 \tab 1 \tab - \tab 4 \tab 2 \tab - \tab 10 \tab 3 \tab - \tab 6 \tab 1 \tab 3 \tab - \tab - \tab 63 \tab 19 \tab 57 \tab - \tab 9 \tab - \tab 1 \tab - \tab - \tab 5 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 4 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1139 \tab 1 \tab 4 \tab 3 \tab 94 \tab 4 \tab 49 \tab 9 \tab 53 \tab 21 \tab 4 \tab 8 \tab 8 \tab - \tab 13 \tab 1 \tab - \tab 4 \tab 57 \tab 3 \tab - \tab 6 \tab 2 \tab 1 \tab 18 \tab 9 \tab 1 \tab 30 \tab 23 \tab - \tab 3 \tab 2 \tab 26 \tab 12 \tab - \tab 95 \tab 18 \tab 22 \tab 1 \tab - \tab 11 \tab 18 \tab 1 \tab 13 \tab 1 \tab 52 \tab 7 \tab 12 \tab - \tab 1 \tab 6 \tab - \tab 1 \tab 1 \tab 5 \tab 6 \tab 3 \tab 10 \tab 7 \tab 2 \tab 6 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 5 \tab 4 \tab - \tab 6 \tab 3 \tab 32 \tab 1 \tab - \tab - \tab 1 \tab 5 \tab 7 \tab 1 \tab 2 \tab 10 \tab 1 \tab 32 \tab 170 \tab 1 \tab 3 \tab 1 \tab 2 \tab 72 \tab 5 \tab 15 \tab 12 \tab 6 \tab - \tab 27 \tab 4 \tab - \tab 4 \tab - \tab 1 \tab 13 \tab 9 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 18 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 9 \tab - \tab 1 \tab 6 \tab 5 \tab 3 \tab 2 \tab - \tab 11 \tab 6 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 13 \tab 6 \tab 4 \tab 1 \tab 9 \tab - \tab 8 \tab 1 \tab 7 \tab - \tab - \tab 11 \tab 5 \tab 5 \tab 1 \tab 3 \tab 4 \tab 3 \tab 1 \tab 1 \tab 5 \tab 6 \tab - \tab 11 \tab 2 \tab 1 \tab 9 \tab - \tab 3 \tab - \tab 1 \tab 64 \tab 11 \tab 39 \tab 1 \tab 6 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1124 \tab 2 \tab 3 \tab 3 \tab 89 \tab 4 \tab 39 \tab 7 \tab 55 \tab 32 \tab 2 \tab 6 \tab 11 \tab 2 \tab 7 \tab 1 \tab 2 \tab 4 \tab 57 \tab 6 \tab - \tab 8 \tab 2 \tab 4 \tab 20 \tab 20 \tab 2 \tab 30 \tab 33 \tab 1 \tab 2 \tab 2 \tab 16 \tab 13 \tab - \tab 72 \tab 15 \tab 23 \tab 2 \tab 1 \tab 12 \tab 13 \tab - \tab 15 \tab 3 \tab 43 \tab 3 \tab 8 \tab 2 \tab 1 \tab 6 \tab 3 \tab - \tab 1 \tab 4 \tab 5 \tab 3 \tab 9 \tab 11 \tab 7 \tab 16 \tab 3 \tab 2 \tab 5 \tab 1 \tab 1 \tab 3 \tab 5 \tab 5 \tab - \tab 1 \tab - \tab 24 \tab - \tab 3 \tab - \tab 1 \tab 7 \tab 14 \tab 1 \tab 3 \tab 9 \tab 1 \tab 37 \tab 135 \tab 3 \tab 3 \tab 2 \tab 1 \tab 78 \tab 6 \tab 8 \tab 15 \tab 6 \tab 1 \tab 28 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 29 \tab 5 \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab 2 \tab - \tab - \tab - \tab 11 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 6 \tab 2 \tab 3 \tab 21 \tab - \tab - \tab 4 \tab 8 \tab 1 \tab 1 \tab - \tab 13 \tab 4 \tab 2 \tab - \tab - \tab 6 \tab 5 \tab - \tab 12 \tab 2 \tab 3 \tab 1 \tab 7 \tab 2 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab 19 \tab - \tab 10 \tab - \tab - \tab 6 \tab 4 \tab 4 \tab - \tab 1 \tab 6 \tab - \tab 17 \tab 4 \tab - \tab 8 \tab - \tab 1 \tab - \tab 1 \tab 92 \tab 24 \tab 44 \tab - \tab 13 \tab - \tab - \tab 1 \tab - \tab 2 \tab 5 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 6 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2690 \tab 8 \tab 3 \tab 5 \tab 110 \tab 6 \tab 56 \tab 9 \tab 53 \tab 32 \tab 1 \tab 7 \tab 11 \tab - \tab 13 \tab 2 \tab 2 \tab 10 \tab 48 \tab 7 \tab - \tab 8 \tab 2 \tab 2 \tab 23 \tab 11 \tab 2 \tab 39 \tab 32 \tab - \tab 1 \tab - \tab 23 \tab 9 \tab 2 \tab 92 \tab 21 \tab 48 \tab 4 \tab - \tab 8 \tab 19 \tab 1 \tab 13 \tab 1 \tab 53 \tab 7 \tab 13 \tab 2 \tab 1 \tab 4 \tab 3 \tab - \tab - \tab 6 \tab - \tab 1 \tab 15 \tab 6 \tab 3 \tab 19 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab 3 \tab 2 \tab 7 \tab - \tab 1 \tab 1 \tab 20 \tab 1 \tab 4 \tab 1 \tab - \tab 5 \tab 9 \tab - \tab 2 \tab 11 \tab 1 \tab 40 \tab 145 \tab - \tab 2 \tab - \tab 1 \tab 72 \tab 9 \tab 13 \tab 16 \tab 6 \tab 2 \tab 34 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab 5 \tab 12 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab 15 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 8 \tab - \tab 2 \tab 2 \tab 3 \tab 3 \tab 1 \tab - \tab 11 \tab 1 \tab 3 \tab 8 \tab 1 \tab 4 \tab 9 \tab 1 \tab 13 \tab 8 \tab 7 \tab - \tab 12 \tab 1 \tab 4 \tab 2 \tab 10 \tab 1 \tab 1 \tab 12 \tab 8 \tab 10 \tab - \tab - \tab 6 \tab 1 \tab 2 \tab 3 \tab 4 \tab 4 \tab - \tab 24 \tab 2 \tab - \tab 7 \tab - \tab 6 \tab - \tab 1 \tab 85 \tab 23 \tab 53 \tab - \tab 16 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab 4 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 730 \tab 1 \tab 4 \tab 2 \tab 80 \tab 4 \tab 31 \tab 9 \tab 47 \tab 24 \tab 3 \tab 9 \tab 9 \tab 1 \tab 11 \tab - \tab 1 \tab 6 \tab 31 \tab 4 \tab - \tab 13 \tab - \tab 2 \tab 9 \tab 13 \tab - \tab 31 \tab 21 \tab - \tab 1 \tab 2 \tab 10 \tab 7 \tab 1 \tab 62 \tab 20 \tab 21 \tab - \tab 1 \tab 8 \tab 14 \tab - \tab 16 \tab 1 \tab 21 \tab 4 \tab 10 \tab 1 \tab - \tab 7 \tab 1 \tab - \tab 1 \tab 6 \tab 4 \tab 2 \tab 8 \tab 4 \tab 3 \tab 19 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 13 \tab - \tab - \tab - \tab 2 \tab 1 \tab 8 \tab 1 \tab 2 \tab 10 \tab 1 \tab 31 \tab 102 \tab 2 \tab - \tab - \tab - \tab 52 \tab 12 \tab 15 \tab 11 \tab 2 \tab 4 \tab 14 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab 3 \tab 7 \tab - \tab - \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 14 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 10 \tab - \tab - \tab 8 \tab 1 \tab 2 \tab 4 \tab 4 \tab 1 \tab 6 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab - \tab 5 \tab 1 \tab 2 \tab 3 \tab 5 \tab 7 \tab 1 \tab 8 \tab 1 \tab 4 \tab - \tab 10 \tab - \tab - \tab 11 \tab 4 \tab 5 \tab - \tab - \tab 4 \tab 2 \tab 3 \tab 1 \tab 2 \tab 4 \tab 1 \tab 14 \tab 1 \tab - \tab 4 \tab - \tab 2 \tab 2 \tab - \tab 49 \tab 17 \tab 39 \tab - \tab 4 \tab 1 \tab - \tab 2 \tab - \tab 6 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 677 \tab 2 \tab 2 \tab 2 \tab 66 \tab - \tab 39 \tab 13 \tab 40 \tab 19 \tab 3 \tab 11 \tab 4 \tab 2 \tab 5 \tab - \tab 4 \tab 8 \tab 52 \tab 3 \tab - \tab 7 \tab 1 \tab 3 \tab 11 \tab 11 \tab - \tab 26 \tab 20 \tab - \tab - \tab 2 \tab 14 \tab 8 \tab 2 \tab 64 \tab 19 \tab 28 \tab 2 \tab - \tab 9 \tab 6 \tab 3 \tab 11 \tab - \tab 38 \tab 3 \tab 8 \tab 2 \tab - \tab 5 \tab 6 \tab 1 \tab 1 \tab 4 \tab 6 \tab - \tab 4 \tab 8 \tab 1 \tab 15 \tab 1 \tab 4 \tab 4 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab 15 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 12 \tab 1 \tab 1 \tab 8 \tab 6 \tab 27 \tab 102 \tab 1 \tab 1 \tab - \tab 2 \tab 45 \tab 12 \tab 17 \tab 6 \tab 7 \tab - \tab 28 \tab - \tab - \tab - \tab - \tab - \tab 6 \tab 14 \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 13 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 13 \tab - \tab 2 \tab 3 \tab 5 \tab - \tab - \tab - \tab 6 \tab 8 \tab 3 \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab 9 \tab 3 \tab 2 \tab 2 \tab 11 \tab 1 \tab 8 \tab - \tab 5 \tab - \tab - \tab 9 \tab 5 \tab 6 \tab 1 \tab 1 \tab 8 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab - \tab 22 \tab 4 \tab - \tab 10 \tab - \tab 2 \tab - \tab - \tab 59 \tab 27 \tab 47 \tab - \tab 6 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 3 \tab 6 \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 647 \tab 2 \tab 4 \tab 6 \tab 74 \tab - \tab 37 \tab 8 \tab 37 \tab 19 \tab 4 \tab 5 \tab 8 \tab 1 \tab 13 \tab 3 \tab 2 \tab 7 \tab 36 \tab 4 \tab - \tab 5 \tab - \tab 2 \tab 25 \tab 18 \tab - \tab 24 \tab 24 \tab - \tab 1 \tab - \tab 18 \tab 6 \tab - \tab 36 \tab 24 \tab 12 \tab 2 \tab - \tab 9 \tab 11 \tab 1 \tab 14 \tab 1 \tab 47 \tab 7 \tab 10 \tab 1 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 6 \tab 13 \tab 5 \tab 5 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab 2 \tab 2 \tab 17 \tab - \tab - \tab 2 \tab - \tab 3 \tab 11 \tab - \tab 1 \tab 12 \tab 3 \tab 34 \tab 85 \tab 2 \tab 2 \tab - \tab 2 \tab 53 \tab 3 \tab 6 \tab 11 \tab 3 \tab - \tab 31 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 5 \tab 11 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 12 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 4 \tab 6 \tab 7 \tab - \tab 1 \tab 5 \tab 5 \tab - \tab 1 \tab - \tab 4 \tab 6 \tab 1 \tab 1 \tab - \tab 4 \tab 2 \tab - \tab 15 \tab 5 \tab 3 \tab - \tab 5 \tab - \tab 8 \tab - \tab 3 \tab - \tab 1 \tab 12 \tab 3 \tab 2 \tab 2 \tab 1 \tab 7 \tab 7 \tab 1 \tab - \tab 1 \tab 5 \tab - \tab 17 \tab 1 \tab - \tab 5 \tab - \tab 4 \tab 1 \tab 4 \tab 46 \tab 21 \tab 44 \tab - \tab 7 \tab 1 \tab - \tab - \tab - \tab 4 \tab 3 \tab 2 \tab 5 \tab 3 \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 2 \tab 5 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 20 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 866 \tab 1 \tab - \tab 4 \tab 78 \tab 2 \tab 36 \tab 11 \tab 46 \tab 27 \tab 4 \tab 9 \tab 11 \tab 3 \tab 9 \tab 1 \tab 1 \tab 5 \tab 42 \tab 4 \tab - \tab 13 \tab 2 \tab 2 \tab 22 \tab 16 \tab - \tab 41 \tab 33 \tab - \tab 3 \tab 2 \tab 12 \tab 6 \tab 4 \tab 73 \tab 24 \tab 19 \tab 1 \tab 1 \tab 19 \tab 18 \tab - \tab 9 \tab 3 \tab 64 \tab 10 \tab 11 \tab 2 \tab 1 \tab 7 \tab 2 \tab 1 \tab - \tab 3 \tab 5 \tab 2 \tab 13 \tab 5 \tab 5 \tab 17 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 4 \tab 3 \tab 4 \tab - \tab 2 \tab 1 \tab 21 \tab - \tab 2 \tab 3 \tab 4 \tab 2 \tab 6 \tab 1 \tab 1 \tab 7 \tab 4 \tab 39 \tab 144 \tab 1 \tab 3 \tab 1 \tab 3 \tab 51 \tab 8 \tab 5 \tab 15 \tab 1 \tab 2 \tab 39 \tab 3 \tab - \tab 5 \tab - \tab 1 \tab 5 \tab 20 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab 10 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 4 \tab 5 \tab 1 \tab 18 \tab 4 \tab - \tab 5 \tab 8 \tab 1 \tab - \tab - \tab 11 \tab 5 \tab 2 \tab 3 \tab - \tab 4 \tab 4 \tab 3 \tab 9 \tab 4 \tab 3 \tab - \tab 12 \tab 1 \tab 8 \tab - \tab 9 \tab - \tab - \tab 10 \tab 6 \tab 9 \tab - \tab - \tab 7 \tab 10 \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab 22 \tab 5 \tab 3 \tab 8 \tab - \tab 4 \tab - \tab 3 \tab 51 \tab 30 \tab 59 \tab - \tab 15 \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab 7 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab - \tab 3 \tab 2 \tab 4 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 875 \tab 1 \tab 3 \tab 6 \tab 128 \tab - \tab 37 \tab 6 \tab 75 \tab 34 \tab 4 \tab 6 \tab 10 \tab - \tab 11 \tab 2 \tab 2 \tab 7 \tab 52 \tab 10 \tab - \tab 18 \tab 1 \tab 3 \tab 26 \tab 22 \tab 2 \tab 45 \tab 40 \tab - \tab 1 \tab 1 \tab 19 \tab 5 \tab 1 \tab 78 \tab 23 \tab 22 \tab 3 \tab 2 \tab 18 \tab 21 \tab - \tab 20 \tab 1 \tab 48 \tab 8 \tab 9 \tab 4 \tab 1 \tab 12 \tab 7 \tab 2 \tab - \tab 4 \tab 3 \tab 3 \tab 15 \tab 4 \tab 3 \tab 13 \tab - \tab 7 \tab 10 \tab - \tab - \tab 4 \tab 4 \tab 4 \tab - \tab 2 \tab - \tab 30 \tab - \tab 4 \tab 1 \tab 1 \tab 8 \tab 19 \tab 4 \tab 2 \tab 10 \tab 3 \tab 41 \tab 125 \tab 3 \tab 5 \tab 1 \tab 1 \tab 55 \tab 12 \tab 9 \tab 17 \tab 3 \tab 2 \tab 31 \tab 3 \tab 1 \tab 4 \tab 1 \tab - \tab 6 \tab 15 \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 25 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 5 \tab 5 \tab 2 \tab 10 \tab 2 \tab 3 \tab 9 \tab 6 \tab 2 \tab 3 \tab - \tab 22 \tab 5 \tab 1 \tab 1 \tab 2 \tab 10 \tab 5 \tab 1 \tab 21 \tab 8 \tab 8 \tab - \tab 16 \tab - \tab 7 \tab 1 \tab 5 \tab - \tab - \tab 19 \tab 6 \tab 16 \tab - \tab - \tab 10 \tab 6 \tab 1 \tab 4 \tab 4 \tab 4 \tab - \tab 20 \tab 2 \tab - \tab 10 \tab 1 \tab 1 \tab 1 \tab 2 \tab 71 \tab 26 \tab 102 \tab 1 \tab 16 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 8 \tab 2 \tab - \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 6 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 4 \tab 2 \tab 6 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 691 \tab 1 \tab - \tab 4 \tab 70 \tab 1 \tab 21 \tab 11 \tab 53 \tab 29 \tab 10 \tab 8 \tab 6 \tab 3 \tab 16 \tab - \tab 1 \tab 5 \tab 37 \tab 3 \tab 1 \tab 4 \tab - \tab 4 \tab 23 \tab 17 \tab 2 \tab 51 \tab 40 \tab - \tab 1 \tab - \tab 10 \tab 8 \tab 2 \tab 75 \tab 23 \tab 6 \tab 1 \tab 1 \tab 19 \tab 12 \tab 2 \tab 20 \tab 2 \tab 46 \tab 2 \tab 11 \tab 2 \tab - \tab 4 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab - \tab 17 \tab 5 \tab 6 \tab 17 \tab 1 \tab 3 \tab 9 \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 14 \tab 1 \tab 5 \tab - \tab 1 \tab 7 \tab 6 \tab - \tab 2 \tab 11 \tab 3 \tab 31 \tab 127 \tab - \tab 4 \tab - \tab 2 \tab 63 \tab 5 \tab 3 \tab 12 \tab 9 \tab 2 \tab 26 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 10 \tab 12 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 16 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 3 \tab 4 \tab 2 \tab 6 \tab - \tab 1 \tab 8 \tab 7 \tab 1 \tab 6 \tab - \tab 23 \tab 7 \tab 3 \tab 4 \tab 1 \tab 2 \tab 3 \tab 3 \tab 24 \tab 6 \tab 7 \tab 2 \tab 12 \tab - \tab 6 \tab 3 \tab 5 \tab - \tab - \tab 16 \tab 2 \tab 20 \tab 1 \tab - \tab 6 \tab 8 \tab - \tab 1 \tab 4 \tab 9 \tab - \tab 32 \tab 4 \tab 1 \tab 3 \tab 2 \tab 5 \tab 2 \tab - \tab 47 \tab 30 \tab 65 \tab - \tab 12 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 5 \tab 2 \tab 1 \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1262 \tab 4 \tab 1 \tab 7 \tab 103 \tab - \tab 60 \tab 15 \tab 81 \tab 34 \tab - \tab 10 \tab 14 \tab 2 \tab 9 \tab 1 \tab 1 \tab 7 \tab 78 \tab 6 \tab - \tab 20 \tab 2 \tab 5 \tab 29 \tab 27 \tab 1 \tab 63 \tab 43 \tab 1 \tab 3 \tab 1 \tab 30 \tab 10 \tab - \tab 89 \tab 49 \tab 28 \tab 3 \tab 2 \tab 28 \tab 30 \tab 1 \tab 12 \tab 2 \tab 81 \tab 11 \tab 12 \tab - \tab 2 \tab 19 \tab 9 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 21 \tab 13 \tab 8 \tab 24 \tab - \tab 3 \tab 13 \tab - \tab - \tab 3 \tab 3 \tab 9 \tab - \tab 2 \tab - \tab 79 \tab - \tab 7 \tab 1 \tab 2 \tab 2 \tab 22 \tab - \tab 4 \tab 8 \tab 6 \tab 80 \tab 234 \tab - \tab 3 \tab 3 \tab 3 \tab 114 \tab 5 \tab 17 \tab 22 \tab 4 \tab 3 \tab 46 \tab - \tab 1 \tab 3 \tab - \tab - \tab 8 \tab 43 \tab 2 \tab - \tab - \tab - \tab 5 \tab 1 \tab 3 \tab 4 \tab 3 \tab 1 \tab 22 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 8 \tab 3 \tab 6 \tab 19 \tab 1 \tab 3 \tab 4 \tab 11 \tab 2 \tab 5 \tab - \tab 17 \tab 9 \tab 6 \tab 3 \tab - \tab 7 \tab 8 \tab 2 \tab 24 \tab 12 \tab 9 \tab 2 \tab 30 \tab 1 \tab 9 \tab 2 \tab 11 \tab - \tab 2 \tab 19 \tab 3 \tab 18 \tab - \tab - \tab 15 \tab 6 \tab 5 \tab 4 \tab 4 \tab 13 \tab 1 \tab 26 \tab 7 \tab 4 \tab 14 \tab - \tab 9 \tab - \tab 1 \tab 146 \tab 30 \tab 115 \tab - \tab 13 \tab - \tab 2 \tab 3 \tab - \tab - \tab 13 \tab 4 \tab 8 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 4 \tab 3 \tab 2 \tab 1 \tab 5 \tab - \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 6 \tab 2 \tab 4 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 4 \tab 4 \tab 7 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 5 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 1039 \tab 2 \tab - \tab 1 \tab 116 \tab 2 \tab 41 \tab 9 \tab 72 \tab 30 \tab - \tab 6 \tab 14 \tab 2 \tab 9 \tab 3 \tab - \tab 8 \tab 60 \tab 6 \tab - \tab 33 \tab 2 \tab 1 \tab 38 \tab 21 \tab - \tab 54 \tab 42 \tab - \tab 3 \tab 2 \tab 24 \tab 15 \tab 1 \tab 86 \tab 31 \tab 39 \tab 1 \tab 2 \tab 34 \tab 21 \tab 1 \tab 21 \tab 1 \tab 66 \tab 11 \tab 11 \tab 1 \tab - \tab 13 \tab 8 \tab - \tab 2 \tab - \tab 4 \tab 2 \tab 12 \tab 15 \tab 11 \tab 18 \tab 2 \tab 2 \tab 8 \tab - \tab - \tab 3 \tab 5 \tab 7 \tab - \tab 4 \tab 1 \tab 55 \tab 2 \tab - \tab 2 \tab 3 \tab 12 \tab 40 \tab 1 \tab 2 \tab 16 \tab 4 \tab 71 \tab 177 \tab 2 \tab 5 \tab - \tab 1 \tab 95 \tab 16 \tab 12 \tab 17 \tab 7 \tab 1 \tab 36 \tab 4 \tab 2 \tab 3 \tab - \tab - \tab 9 \tab 38 \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 5 \tab 1 \tab - \tab 15 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 7 \tab 2 \tab - \tab 18 \tab - \tab 2 \tab 8 \tab 9 \tab 1 \tab 3 \tab - \tab 17 \tab 9 \tab 1 \tab 2 \tab - \tab 11 \tab 6 \tab 3 \tab 12 \tab 10 \tab 3 \tab 3 \tab 27 \tab - \tab 9 \tab 1 \tab 12 \tab - \tab - \tab 29 \tab 5 \tab 23 \tab - \tab 3 \tab 12 \tab 9 \tab 2 \tab 1 \tab 4 \tab 9 \tab 1 \tab 35 \tab 4 \tab - \tab 6 \tab - \tab 6 \tab - \tab 1 \tab 115 \tab 39 \tab 110 \tab - \tab 11 \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab 6 \tab 2 \tab 9 \tab - \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab 3 \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 3 \tab - \tab 4 \tab - \tab 4 \tab - \tab 8 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 2 \tab 5 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab 8 \tab - \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 3544 \cr
   Total \tab 46034 \tab 32 \tab 61 \tab 74 \tab 1408 \tab 46 \tab 706 \tab 136 \tab 809 \tab 397 \tab 45 \tab 128 \tab 156 \tab 20 \tab 168 \tab 18 \tab 22 \tab 90 \tab 733 \tab 69 \tab 3 \tab 151 \tab 15 \tab 45 \tab 305 \tab 227 \tab 20 \tab 599 \tab 447 \tab 4 \tab 21 \tab 22 \tab 279 \tab 134 \tab 17 \tab 1119 \tab 361 \tab 387 \tab 28 \tab 11 \tab 216 \tab 240 \tab 15 \tab 253 \tab 20 \tab 766 \tab 92 \tab 164 \tab 32 \tab 14 \tab 114 \tab 54 \tab 6 \tab 8 \tab 67 \tab 48 \tab 28 \tab 167 \tab 117 \tab 67 \tab 234 \tab 14 \tab 41 \tab 73 \tab 7 \tab 14 \tab 36 \tab 51 \tab 67 \tab 1 \tab 34 \tab 9 \tab 394 \tab 7 \tab 31 \tab 13 \tab 19 \tab 67 \tab 186 \tab 18 \tab 29 \tab 162 \tab 41 \tab 657 \tab 2112 \tab 19 \tab 39 \tab 10 \tab 28 \tab 1013 \tab 139 \tab 182 \tab 193 \tab 72 \tab 26 \tab 471 \tab 27 \tab 13 \tab 40 \tab 6 \tab 4 \tab 112 \tab 292 \tab 28 \tab 2 \tab 1 \tab 8 \tab 30 \tab 18 \tab 30 \tab 24 \tab 13 \tab 2 \tab 225 \tab 11 \tab 12 \tab 12 \tab 1 \tab 11 \tab 5 \tab 63 \tab 32 \tab 25 \tab 184 \tab 11 \tab 21 \tab 68 \tab 82 \tab 20 \tab 39 \tab 1 \tab 167 \tab 86 \tab 30 \tab 40 \tab 8 \tab 64 \tab 57 \tab 23 \tab 202 \tab 81 \tab 75 \tab 15 \tab 188 \tab 11 \tab 98 \tab 12 \tab 91 \tab 2 \tab 10 \tab 213 \tab 57 \tab 153 \tab 6 \tab 9 \tab 103 \tab 70 \tab 25 \tab 21 \tab 39 \tab 80 \tab 5 \tab 296 \tab 41 \tab 11 \tab 104 \tab 6 \tab 60 \tab 7 \tab 17 \tab 1094 \tab 367 \tab 945 \tab 3 \tab 145 \tab 16 \tab 16 \tab 11 \tab 4 \tab 51 \tab 47 \tab 21 \tab 66 \tab 13 \tab 5 \tab 9 \tab 4 \tab 43 \tab 15 \tab 5 \tab 6 \tab 3 \tab 14 \tab 8 \tab 12 \tab 20 \tab 16 \tab 7 \tab 13 \tab 12 \tab 23 \tab 30 \tab 12 \tab 6 \tab 24 \tab 3 \tab 10 \tab 14 \tab 28 \tab 12 \tab 15 \tab 4 \tab 20 \tab 10 \tab 51 \tab 4 \tab 14 \tab 3 \tab 12 \tab 18 \tab 3 \tab 20 \tab 15 \tab 17 \tab 1 \tab 2 \tab 3 \tab 4 \tab 7 \tab 4 \tab 8 \tab 7 \tab 4 \tab 6 \tab 27 \tab 19 \tab 55 \tab 4 \tab 11 \tab 12 \tab 4 \tab 5 \tab 5 \tab 9 \tab 1 \tab 9 \tab 3 \tab 15 \tab 6 \tab 4 \tab 3 \tab 7 \tab 3 \tab 15 \tab 9 \tab 7 \tab 3 \tab 4 \tab 3 \tab 10 \tab 6 \tab 9 \tab 3 \tab 7 \tab 5 \tab 4 \tab 11 \tab 7 \tab 8 \tab 23 \tab 3 \tab 20 \tab 7 \tab 9 \tab 1 \tab 2 \tab 3 \tab 3 \tab 3 \tab 2 \tab 7 \tab 11 \tab 6 \tab 4 \tab 6 \tab 3 \tab 4 \tab 13 \tab 1 \tab 6 \tab 4 \tab 3 \tab 4 \tab 2 \tab 1 \tab 1 \tab 24 \tab 11 \tab 1 \tab 12 \tab 1 \tab 1 \tab 3 \tab 5 \tab 2 \tab 1 \tab 3 \tab 5 \tab 2 \tab 2 \tab 3 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} i don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
