% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_598.R
\name{masei10inc}
\alias{masei10inc}
\title{Percentage of $45k+ earners in maocc10 based on acs 2010}
\source{
General Social Survey https://gss.norc.org
}
\description{
masei10inc
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1.5 \tab 11.3 \tab 11.5 \tab 11.6 \tab 11.7 \tab 12.2 \tab 12.7 \tab 13.2 \tab 13.5 \tab 13.7 \tab 13.8 \tab 13.9 \tab 14.7 \tab 14.9 \tab 15.1 \tab 15.6 \tab 15.9 \tab 16.8 \tab 16.9 \tab 17.3 \tab 17.4 \tab 17.7 \tab 17.9 \tab 18.5 \tab 18.8 \tab 18.9 \tab 19 \tab 19.3 \tab 19.7 \tab 19.9 \tab 20.6 \tab 21.1 \tab 21.2 \tab 21.4 \tab 21.5 \tab 21.6 \tab 22.1 \tab 22.2 \tab 22.4 \tab 22.8 \tab 23.1 \tab 23.4 \tab 23.6 \tab 23.8 \tab 24.3 \tab 24.7 \tab 25.4 \tab 25.5 \tab 25.7 \tab 25.9 \tab 26.4 \tab 26.6 \tab 26.7 \tab 26.9 \tab 27 \tab 27.8 \tab 27.9 \tab 28.2 \tab 28.7 \tab 29.3 \tab 3.2 \tab 3.3 \tab 3.5 \tab 3.6 \tab 3.8 \tab 30.9 \tab 31.1 \tab 31.4 \tab 31.5 \tab 31.7 \tab 32.4 \tab 33 \tab 33.4 \tab 33.6 \tab 33.8 \tab 35 \tab 35.3 \tab 35.8 \tab 35.9 \tab 37.4 \tab 38.5 \tab 39.1 \tab 39.4 \tab 4.1 \tab 4.3 \tab 4.5 \tab 4.6 \tab 40.3 \tab 40.4 \tab 41.1 \tab 41.5 \tab 42.7 \tab 42.8 \tab 43.4 \tab 43.6 \tab 43.7 \tab 44.2 \tab 44.5 \tab 45.2 \tab 46.5 \tab 46.9 \tab 47.6 \tab 48.2 \tab 49.6 \tab 5 \tab 5.2 \tab 5.5 \tab 50.4 \tab 50.5 \tab 51.8 \tab 52.4 \tab 53.2 \tab 53.7 \tab 54.2 \tab 54.3 \tab 54.7 \tab 54.9 \tab 55 \tab 55.1 \tab 55.2 \tab 55.3 \tab 56.1 \tab 56.5 \tab 58.5 \tab 58.6 \tab 59.3 \tab 59.4 \tab 59.7 \tab 59.9 \tab 6.4 \tab 6.9 \tab 60.1 \tab 60.7 \tab 60.9 \tab 62.1 \tab 62.3 \tab 62.4 \tab 62.9 \tab 64.2 \tab 64.3 \tab 64.5 \tab 64.7 \tab 64.9 \tab 65.3 \tab 65.6 \tab 66.2 \tab 67.1 \tab 67.5 \tab 67.6 \tab 68 \tab 68.1 \tab 68.4 \tab 69.6 \tab 7.1 \tab 7.2 \tab 7.5 \tab 7.7 \tab 7.8 \tab 70.3 \tab 71.2 \tab 72.2 \tab 72.3 \tab 72.5 \tab 74 \tab 74.2 \tab 74.7 \tab 75.6 \tab 76.4 \tab 76.8 \tab 77.8 \tab 78.4 \tab 79.5 \tab 79.6 \tab 79.7 \tab 8.8 \tab 8.9 \tab 80.3 \tab 82.4 \tab 84.3 \tab 85.6 \tab 86 \tab 87 \tab 89.6 \tab 9 \tab 9.4 \tab 9.5 \tab 9.9 \tab 91.5 \tab 93.9 \tab 94.4 \tab 10.3 \tab 11.8 \tab 12.5 \tab 13.4 \tab 14.5 \tab 15.3 \tab 16.4 \tab 17.1 \tab 17.6 \tab 21.9 \tab 22 \tab 26.8 \tab 28.8 \tab 32.5 \tab 35.7 \tab 41.4 \tab 44.8 \tab 48.5 \tab 52.6 \tab 63.1 \tab 63.6 \tab 65 \tab 67.7 \tab 69.5 \tab 73.9 \tab 74.6 \tab 75 \tab 78.9 \tab 81.1 \tab 82.6 \tab 84.2 \tab 84.7 \tab 88.1 \tab 90.5 \tab 93.2 \tab 93.3 \tab 95.5 \tab 10.9 \tab 12.8 \tab 20.7 \tab 26.1 \tab 30.3 \tab 30.7 \tab 30.8 \tab 34.1 \tab 42.2 \tab 42.5 \tab 43 \tab 44.7 \tab 47.5 \tab 51.4 \tab 60 \tab 61 \tab 61.1 \tab 61.4 \tab 63 \tab 64.8 \tab 66.4 \tab 7.9 \tab 77.6 \tab 82.3 \tab 86.1 \tab 10.7 \tab 17.2 \tab 18.6 \tab 20.2 \tab 23.5 \tab 23.9 \tab 29.5 \tab 33.3 \tab 37 \tab 39.7 \tab 60.8 \tab 63.2 \tab 70.4 \tab 77.3 \tab 80.4 \tab 81.7 \tab 81.8 \tab 81.9 \tab 83.1 \tab 86.9 \tab 87.6 \tab 90.2 \tab 91.9 \tab 92.7 \tab 93.4 \tab 94 \tab 26 \tab 34 \tab 37.3 \tab 40 \tab 40.8 \tab 41.9 \tab 57.3 \tab 60.2 \tab 66 \tab 79.8 \tab 87.9 \tab 90.1 \tab 92.9 \tab 12 \tab 29.8 \tab 33.9 \tab 41.6 \tab 45.5 \tab 53.1 \tab 57.2 \tab 61.3 \tab 70.5 \tab 71 \tab 71.9 \tab 72 \tab 82 \tab 91.1 \tab 15.7 \tab 59.2 \tab 17 \tab 24.9 \tab 30.4 \tab 50.7 \tab 51.6 \tab 57.1 \tab 60.6 \tab 59.8 \tab 7 \tab 71.3 \tab 73.7 \tab 57.4 \tab 75.4 \tab 89.8 \tab 91.7 \tab 34.5 \tab 42.1 \tab 46.4 \tab 47.7 \tab 87.2 \tab 26.5 \tab 72.7 \tab 91.4 \tab 20.5 \tab 22.6 \tab 45.3 \tab 79 \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1351 \tab 3 \tab 73 \tab 14 \tab 1 \tab 1 \tab 6 \tab 4 \tab 26 \tab 1 \tab 2 \tab 11 \tab 2 \tab 19 \tab 12 \tab 1 \tab 2 \tab 21 \tab 21 \tab 1 \tab 3 \tab 6 \tab 3 \tab 1 \tab 3 \tab 8 \tab 9 \tab 5 \tab 1 \tab 2 \tab 2 \tab 1 \tab 7 \tab 2 \tab 1 \tab 2 \tab 32 \tab 3 \tab 6 \tab 2 \tab 18 \tab 2 \tab 3 \tab 7 \tab 1 \tab 6 \tab 3 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 39 \tab 1 \tab 5 \tab 138 \tab 51 \tab 7 \tab 3 \tab 1 \tab 6 \tab 3 \tab 1 \tab 5 \tab 5 \tab 75 \tab 1 \tab 1 \tab 4 \tab 2 \tab 3 \tab 4 \tab 1 \tab 1 \tab 3 \tab 23 \tab 13 \tab 2 \tab 1 \tab 1 \tab 18 \tab 36 \tab 1 \tab 45 \tab 7 \tab 106 \tab 7 \tab 11 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 3 \tab 1 \tab 1 \tab 9 \tab 4 \tab 1 \tab 1 \tab 3 \tab 4 \tab 9 \tab 51 \tab 15 \tab 67 \tab 2 \tab 5 \tab 12 \tab 1 \tab 9 \tab 2 \tab 1 \tab 1 \tab 3 \tab 5 \tab 5 \tab 5 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 12 \tab 7 \tab 23 \tab 2 \tab 62 \tab 1 \tab 1 \tab 1 \tab 9 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 18 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab 15 \tab 19 \tab 6 \tab 2 \tab 1 \tab 5 \tab 15 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 6 \tab 16 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 52 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 14 \tab 8 \tab 1 \tab 46 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1158 \tab 5 \tab 80 \tab 19 \tab - \tab 1 \tab 3 \tab 7 \tab 24 \tab 2 \tab 1 \tab 10 \tab - \tab 16 \tab 20 \tab - \tab 2 \tab 14 \tab 17 \tab - \tab 8 \tab 3 \tab 2 \tab - \tab 5 \tab 8 \tab 9 \tab 7 \tab 4 \tab - \tab 2 \tab 4 \tab 11 \tab 1 \tab - \tab 3 \tab 30 \tab 3 \tab 5 \tab 3 \tab 15 \tab 2 \tab 3 \tab 7 \tab - \tab 6 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab 45 \tab - \tab 9 \tab 151 \tab 57 \tab 7 \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 2 \tab 14 \tab 65 \tab - \tab 3 \tab 9 \tab - \tab 3 \tab 8 \tab 1 \tab - \tab 5 \tab 22 \tab 14 \tab - \tab - \tab - \tab 14 \tab 28 \tab 2 \tab 44 \tab 12 \tab 105 \tab 9 \tab 13 \tab - \tab 1 \tab - \tab - \tab 2 \tab 7 \tab 10 \tab - \tab 1 \tab 5 \tab 4 \tab 2 \tab 1 \tab 1 \tab 6 \tab 14 \tab 50 \tab 4 \tab 72 \tab - \tab 5 \tab 18 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 6 \tab 7 \tab 6 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab 12 \tab 11 \tab 23 \tab 3 \tab 71 \tab 1 \tab 1 \tab - \tab 9 \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab 26 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 9 \tab 18 \tab 6 \tab 1 \tab 2 \tab 8 \tab 18 \tab 1 \tab 5 \tab 4 \tab 1 \tab 2 \tab 1 \tab 9 \tab - \tab 5 \tab 1 \tab 6 \tab 1 \tab 3 \tab 3 \tab 57 \tab - \tab - \tab - \tab 1 \tab 9 \tab - \tab 21 \tab 10 \tab 4 \tab 43 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1213 \tab 10 \tab 74 \tab 16 \tab - \tab - \tab 2 \tab 13 \tab 20 \tab - \tab 1 \tab 9 \tab - \tab 13 \tab 15 \tab - \tab - \tab 21 \tab 19 \tab 4 \tab 5 \tab 1 \tab 3 \tab - \tab 3 \tab 4 \tab 8 \tab 5 \tab 4 \tab - \tab 1 \tab 1 \tab 7 \tab 3 \tab - \tab 1 \tab 31 \tab 1 \tab 3 \tab 2 \tab 23 \tab 2 \tab 1 \tab 4 \tab - \tab 10 \tab 2 \tab 5 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 50 \tab 2 \tab 7 \tab 125 \tab 38 \tab 8 \tab 4 \tab 2 \tab 2 \tab 2 \tab - \tab 8 \tab 5 \tab 60 \tab 1 \tab 2 \tab 4 \tab 3 \tab 2 \tab 6 \tab 3 \tab - \tab 2 \tab 17 \tab 18 \tab 1 \tab - \tab - \tab 10 \tab 32 \tab 1 \tab 42 \tab 11 \tab 87 \tab 4 \tab 10 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 15 \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab 7 \tab 10 \tab 42 \tab 13 \tab 68 \tab - \tab 1 \tab 12 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 9 \tab 5 \tab 7 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 6 \tab 1 \tab 1 \tab 8 \tab 7 \tab 31 \tab - \tab 74 \tab - \tab - \tab - \tab 9 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 26 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 4 \tab 3 \tab - \tab - \tab 14 \tab 14 \tab 16 \tab 1 \tab - \tab 5 \tab 13 \tab - \tab 7 \tab 7 \tab 2 \tab 4 \tab - \tab 9 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 2 \tab 2 \tab 62 \tab - \tab 1 \tab - \tab - \tab 5 \tab - \tab 6 \tab 14 \tab 6 \tab 44 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 9 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1184 \tab 5 \tab 71 \tab 18 \tab - \tab 2 \tab 5 \tab 8 \tab 21 \tab 2 \tab 1 \tab 5 \tab - \tab 13 \tab 13 \tab - \tab 1 \tab 16 \tab 29 \tab 2 \tab 4 \tab 2 \tab 3 \tab - \tab 3 \tab 3 \tab 2 \tab 6 \tab 2 \tab - \tab 3 \tab 2 \tab 5 \tab - \tab 1 \tab - \tab 29 \tab 1 \tab 2 \tab 6 \tab 18 \tab - \tab 2 \tab 11 \tab - \tab 8 \tab 1 \tab - \tab - \tab 6 \tab 1 \tab 1 \tab - \tab 44 \tab 2 \tab 6 \tab 151 \tab 41 \tab 11 \tab 4 \tab - \tab 3 \tab 3 \tab 5 \tab 5 \tab 5 \tab 63 \tab - \tab 2 \tab 5 \tab 2 \tab - \tab 5 \tab 3 \tab 2 \tab 5 \tab 19 \tab 17 \tab 1 \tab - \tab - \tab 8 \tab 35 \tab 3 \tab 51 \tab 10 \tab 102 \tab 3 \tab 11 \tab 1 \tab 2 \tab - \tab 4 \tab 3 \tab 4 \tab 11 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 11 \tab 14 \tab 40 \tab 5 \tab 57 \tab - \tab 1 \tab 12 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 5 \tab 6 \tab 5 \tab 8 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 6 \tab 2 \tab 2 \tab 9 \tab 6 \tab 31 \tab - \tab 63 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 3 \tab 4 \tab - \tab 1 \tab 19 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 12 \tab 16 \tab 15 \tab 1 \tab - \tab 9 \tab 10 \tab - \tab 7 \tab 3 \tab - \tab 4 \tab 1 \tab 12 \tab 2 \tab 3 \tab 1 \tab 9 \tab 3 \tab 2 \tab 1 \tab 57 \tab - \tab 1 \tab - \tab - \tab 5 \tab - \tab 9 \tab 4 \tab 3 \tab 44 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 4 \tab 6 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1139 \tab 2 \tab 95 \tab 17 \tab - \tab 1 \tab 1 \tab 3 \tab 18 \tab - \tab - \tab 12 \tab 1 \tab 18 \tab 14 \tab 2 \tab - \tab 31 \tab 13 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 7 \tab 7 \tab 3 \tab 3 \tab - \tab 2 \tab 1 \tab 10 \tab 1 \tab - \tab 1 \tab 24 \tab 1 \tab - \tab 6 \tab 32 \tab 2 \tab - \tab 5 \tab - \tab 7 \tab 1 \tab 3 \tab - \tab 5 \tab 2 \tab - \tab - \tab 51 \tab 1 \tab 5 \tab 170 \tab 29 \tab 10 \tab 2 \tab - \tab 6 \tab 4 \tab 1 \tab 6 \tab 9 \tab 72 \tab 1 \tab 1 \tab 5 \tab 4 \tab 1 \tab 8 \tab 4 \tab - \tab - \tab 9 \tab 15 \tab - \tab 3 \tab 1 \tab 10 \tab 27 \tab 1 \tab 47 \tab 8 \tab 94 \tab 4 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 6 \tab 9 \tab - \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab 11 \tab 9 \tab 57 \tab 4 \tab 49 \tab - \tab - \tab 18 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 5 \tab 4 \tab 12 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 9 \tab 1 \tab - \tab 13 \tab 3 \tab 29 \tab 1 \tab 64 \tab - \tab 1 \tab - \tab 9 \tab 1 \tab - \tab - \tab 2 \tab 6 \tab 1 \tab 1 \tab 11 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 6 \tab - \tab 2 \tab 3 \tab - \tab 16 \tab 13 \tab 8 \tab - \tab 4 \tab 9 \tab 11 \tab - \tab 5 \tab 6 \tab - \tab 3 \tab 3 \tab 11 \tab 3 \tab 2 \tab 2 \tab 5 \tab 5 \tab 2 \tab - \tab 39 \tab - \tab - \tab 1 \tab 1 \tab 8 \tab 1 \tab 8 \tab 11 \tab 4 \tab 36 \tab 2 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1124 \tab 2 \tab 72 \tab 13 \tab 1 \tab 1 \tab 4 \tab 3 \tab 14 \tab 3 \tab 1 \tab 13 \tab 1 \tab 20 \tab 10 \tab 2 \tab 1 \tab 18 \tab 13 \tab 3 \tab 7 \tab - \tab 1 \tab - \tab 5 \tab 7 \tab 8 \tab 5 \tab 7 \tab 1 \tab 2 \tab 3 \tab 6 \tab 1 \tab - \tab 3 \tab 24 \tab 1 \tab 3 \tab 6 \tab 24 \tab 2 \tab 1 \tab 6 \tab - \tab 14 \tab 2 \tab - \tab - \tab 7 \tab 2 \tab - \tab 1 \tab 42 \tab - \tab 4 \tab 135 \tab 34 \tab 9 \tab 1 \tab 4 \tab 8 \tab 2 \tab 2 \tab 7 \tab 7 \tab 78 \tab 2 \tab 5 \tab 5 \tab 1 \tab 1 \tab 10 \tab - \tab - \tab 2 \tab 28 \tab 8 \tab - \tab 1 \tab - \tab 9 \tab 28 \tab 4 \tab 47 \tab 19 \tab 89 \tab 4 \tab 5 \tab 1 \tab 1 \tab - \tab 3 \tab 6 \tab 1 \tab 7 \tab 1 \tab 2 \tab 7 \tab 1 \tab 2 \tab 1 \tab 4 \tab 12 \tab 21 \tab 57 \tab 3 \tab 41 \tab - \tab 5 \tab 11 \tab 5 \tab - \tab - \tab 3 \tab 4 \tab 7 \tab 13 \tab 6 \tab 6 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 8 \tab - \tab 2 \tab 1 \tab 3 \tab 40 \tab 1 \tab 92 \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 3 \tab 4 \tab - \tab - \tab 24 \tab - \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab - \tab 2 \tab - \tab 13 \tab 23 \tab 6 \tab 1 \tab - \tab 13 \tab 17 \tab - \tab 7 \tab 6 \tab 1 \tab 4 \tab - \tab 19 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab 3 \tab 44 \tab - \tab - \tab - \tab 2 \tab 5 \tab - \tab 11 \tab 11 \tab 6 \tab 31 \tab 5 \tab 1 \tab 4 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2690 \tab 2 \tab 92 \tab 16 \tab 2 \tab 2 \tab 2 \tab 5 \tab 20 \tab 4 \tab 1 \tab 9 \tab - \tab 23 \tab 22 \tab 2 \tab - \tab 23 \tab 13 \tab 2 \tab 6 \tab 1 \tab 1 \tab - \tab 3 \tab 4 \tab 14 \tab 8 \tab 5 \tab - \tab 2 \tab 1 \tab 8 \tab 1 \tab - \tab 1 \tab 48 \tab - \tab 5 \tab 1 \tab 20 \tab - \tab 2 \tab 8 \tab - \tab 9 \tab 4 \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 49 \tab 2 \tab 6 \tab 145 \tab 39 \tab 13 \tab 2 \tab 2 \tab 3 \tab 1 \tab 8 \tab 5 \tab 9 \tab 72 \tab - \tab 1 \tab 4 \tab 4 \tab 1 \tab 9 \tab 2 \tab 1 \tab 2 \tab 11 \tab 13 \tab 1 \tab 1 \tab 1 \tab 11 \tab 34 \tab 2 \tab 50 \tab 11 \tab 110 \tab 6 \tab 8 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 3 \tab 12 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 11 \tab 8 \tab 48 \tab 3 \tab 56 \tab - \tab 9 \tab 15 \tab 4 \tab 8 \tab - \tab 1 \tab 2 \tab 3 \tab 11 \tab 10 \tab 3 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 7 \tab - \tab 1 \tab 5 \tab 5 \tab 39 \tab 1 \tab 85 \tab - \tab 1 \tab 1 \tab 7 \tab 3 \tab - \tab - \tab 5 \tab 2 \tab - \tab 1 \tab 23 \tab 2 \tab 2 \tab - \tab - \tab - \tab 8 \tab 1 \tab 2 \tab 1 \tab 2 \tab 17 \tab 25 \tab 7 \tab 1 \tab 6 \tab 16 \tab 24 \tab - \tab 9 \tab 4 \tab 3 \tab 1 \tab 4 \tab 12 \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab 2 \tab 3 \tab 53 \tab - \tab - \tab - \tab 1 \tab 11 \tab - \tab 11 \tab 7 \tab 7 \tab 42 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 4 \tab 3 \tab - \tab 4 \tab - \tab - \tab 4 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 730 \tab 6 \tab 62 \tab 16 \tab - \tab - \tab 2 \tab 4 \tab 19 \tab - \tab 2 \tab 7 \tab 1 \tab 9 \tab 13 \tab - \tab 1 \tab 17 \tab 16 \tab 2 \tab 7 \tab 2 \tab - \tab - \tab - \tab 4 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 4 \tab 1 \tab - \tab 1 \tab 23 \tab 2 \tab 7 \tab 5 \tab 13 \tab 1 \tab 4 \tab 5 \tab - \tab 8 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 17 \tab 1 \tab 6 \tab 102 \tab 29 \tab 8 \tab - \tab - \tab 5 \tab 3 \tab 1 \tab 7 \tab 9 \tab 52 \tab - \tab 2 \tab 2 \tab 5 \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab 7 \tab 15 \tab 1 \tab - \tab - \tab 10 \tab 14 \tab 2 \tab 42 \tab 13 \tab 80 \tab 4 \tab 10 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab 3 \tab 8 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 4 \tab 2 \tab 8 \tab 31 \tab 4 \tab 32 \tab - \tab 2 \tab 14 \tab 5 \tab - \tab 1 \tab - \tab 3 \tab 4 \tab 14 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 3 \tab 2 \tab 33 \tab 1 \tab 49 \tab - \tab 1 \tab - \tab 7 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab - \tab 2 \tab 17 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab 14 \tab 14 \tab 9 \tab - \tab 3 \tab 4 \tab 14 \tab 1 \tab 6 \tab 5 \tab 2 \tab 2 \tab 2 \tab 11 \tab 1 \tab 6 \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab 39 \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 9 \tab 6 \tab 4 \tab 34 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 677 \tab 2 \tab 66 \tab 8 \tab 1 \tab - \tab 3 \tab 5 \tab 19 \tab 1 \tab - \tab 8 \tab 1 \tab 11 \tab 15 \tab 1 \tab 3 \tab 16 \tab 11 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 4 \tab 7 \tab 3 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 29 \tab 2 \tab 3 \tab 6 \tab 15 \tab - \tab - \tab 7 \tab - \tab 12 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 33 \tab 1 \tab 4 \tab 102 \tab 30 \tab 3 \tab 2 \tab 1 \tab 5 \tab 3 \tab 2 \tab 7 \tab 13 \tab 45 \tab - \tab - \tab 4 \tab - \tab 6 \tab 8 \tab - \tab 1 \tab 2 \tab 13 \tab 17 \tab - \tab 1 \tab - \tab 8 \tab 28 \tab 1 \tab 35 \tab 9 \tab 66 \tab - \tab 12 \tab - \tab 2 \tab 1 \tab 2 \tab 4 \tab 4 \tab 11 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab - \tab 6 \tab 13 \tab 52 \tab 2 \tab 41 \tab - \tab 2 \tab 13 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 5 \tab 9 \tab 8 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 10 \tab - \tab 1 \tab 6 \tab 2 \tab 25 \tab 1 \tab 59 \tab 1 \tab - \tab - \tab 8 \tab - \tab - \tab - \tab 3 \tab 3 \tab - \tab 1 \tab 27 \tab 2 \tab 4 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 6 \tab 11 \tab 11 \tab - \tab - \tab 7 \tab 22 \tab - \tab 7 \tab 3 \tab 1 \tab 3 \tab 4 \tab 9 \tab - \tab - \tab - \tab 7 \tab 1 \tab 1 \tab 5 \tab 47 \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 4 \tab 9 \tab 3 \tab 27 \tab 3 \tab 3 \tab 6 \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 5 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 647 \tab 4 \tab 36 \tab 15 \tab 2 \tab 3 \tab 2 \tab 3 \tab 23 \tab - \tab 1 \tab 6 \tab - \tab 25 \tab 14 \tab - \tab - \tab 14 \tab 14 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 13 \tab 12 \tab 4 \tab 3 \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 12 \tab - \tab 3 \tab 2 \tab 17 \tab 2 \tab - \tab 1 \tab - \tab 11 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 47 \tab 1 \tab 3 \tab 85 \tab 33 \tab 5 \tab 3 \tab - \tab 3 \tab 4 \tab 2 \tab 4 \tab 8 \tab 53 \tab - \tab 4 \tab 2 \tab 2 \tab 3 \tab 6 \tab 2 \tab - \tab 1 \tab 11 \tab 6 \tab - \tab 2 \tab 1 \tab 12 \tab 31 \tab 5 \tab 34 \tab 17 \tab 74 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 3 \tab 5 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 4 \tab 7 \tab 36 \tab 4 \tab 38 \tab - \tab 2 \tab 12 \tab 4 \tab 1 \tab - \tab 6 \tab 1 \tab 5 \tab 10 \tab 7 \tab 8 \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab 5 \tab 3 \tab - \tab 5 \tab 6 \tab 28 \tab 3 \tab 46 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab 1 \tab 5 \tab 1 \tab - \tab 21 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab - \tab 10 \tab 11 \tab 5 \tab 1 \tab - \tab 8 \tab 17 \tab - \tab 2 \tab 5 \tab 1 \tab 7 \tab 1 \tab 12 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 4 \tab 3 \tab 44 \tab - \tab - \tab - \tab 4 \tab 9 \tab - \tab 8 \tab 8 \tab 5 \tab 28 \tab 3 \tab 2 \tab 5 \tab - \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 20 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 866 \tab 1 \tab 73 \tab 14 \tab - \tab 1 \tab 2 \tab 3 \tab 24 \tab 2 \tab 2 \tab 6 \tab - \tab 22 \tab 15 \tab 2 \tab 1 \tab 19 \tab 9 \tab 1 \tab 7 \tab 2 \tab - \tab 1 \tab 2 \tab 5 \tab 7 \tab 1 \tab 1 \tab - \tab 5 \tab 6 \tab 6 \tab 1 \tab - \tab 1 \tab 23 \tab 2 \tab 3 \tab 6 \tab 21 \tab 1 \tab 2 \tab 7 \tab - \tab 6 \tab 1 \tab 4 \tab 1 \tab 2 \tab 3 \tab 4 \tab 1 \tab 62 \tab 2 \tab 3 \tab 144 \tab 33 \tab 11 \tab 4 \tab - \tab 1 \tab 4 \tab 1 \tab 9 \tab 11 \tab 51 \tab 1 \tab 1 \tab 4 \tab 5 \tab 4 \tab 3 \tab 3 \tab 1 \tab 2 \tab 20 \tab 5 \tab - \tab 1 \tab 1 \tab 7 \tab 39 \tab 4 \tab 45 \tab 16 \tab 78 \tab 2 \tab 8 \tab - \tab 2 \tab - \tab - \tab 3 \tab 4 \tab 12 \tab - \tab 1 \tab 6 \tab 1 \tab 4 \tab 1 \tab - \tab 11 \tab 18 \tab 42 \tab - \tab 39 \tab - \tab 4 \tab 10 \tab 2 \tab 3 \tab 4 \tab 1 \tab 2 \tab 8 \tab 11 \tab 6 \tab 3 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab 8 \tab 2 \tab 1 \tab 5 \tab 4 \tab 45 \tab 2 \tab 51 \tab - \tab - \tab - \tab 10 \tab 1 \tab - \tab - \tab 2 \tab 5 \tab 1 \tab 3 \tab 30 \tab - \tab 5 \tab 1 \tab - \tab - \tab 4 \tab 3 \tab - \tab 3 \tab 4 \tab 15 \tab 17 \tab 9 \tab - \tab 2 \tab 15 \tab 22 \tab - \tab 6 \tab 3 \tab 1 \tab 10 \tab 4 \tab 10 \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 59 \tab 2 \tab - \tab - \tab 3 \tab 8 \tab 1 \tab 11 \tab 17 \tab 4 \tab 51 \tab 3 \tab - \tab 7 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 6 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 875 \tab 4 \tab 80 \tab 19 \tab 2 \tab 2 \tab 3 \tab 2 \tab 22 \tab 4 \tab 2 \tab 5 \tab - \tab 26 \tab 14 \tab 1 \tab - \tab 30 \tab 20 \tab 2 \tab 3 \tab 3 \tab 1 \tab 2 \tab 10 \tab 4 \tab 5 \tab 6 \tab 8 \tab - \tab 1 \tab 3 \tab 6 \tab - \tab - \tab 1 \tab 23 \tab 1 \tab 5 \tab 4 \tab 31 \tab 2 \tab 2 \tab 5 \tab - \tab 19 \tab 3 \tab - \tab 1 \tab 3 \tab 5 \tab 2 \tab 2 \tab 50 \tab 2 \tab 4 \tab 125 \tab 39 \tab 15 \tab 2 \tab - \tab 3 \tab 4 \tab 1 \tab 9 \tab 6 \tab 55 \tab 1 \tab - \tab 10 \tab 4 \tab 3 \tab 7 \tab 3 \tab 1 \tab 4 \tab 14 \tab 9 \tab 1 \tab - \tab 1 \tab 10 \tab 31 \tab 2 \tab 72 \tab 17 \tab 128 \tab - \tab 10 \tab - \tab - \tab 1 \tab - \tab 5 \tab 6 \tab 16 \tab - \tab 2 \tab 4 \tab 4 \tab 1 \tab 1 \tab 2 \tab 20 \tab 10 \tab 52 \tab 3 \tab 37 \tab - \tab 5 \tab 25 \tab 10 \tab 1 \tab 2 \tab 2 \tab 1 \tab 6 \tab 15 \tab 16 \tab 5 \tab 2 \tab 2 \tab - \tab 5 \tab - \tab - \tab 10 \tab 2 \tab - \tab 6 \tab 6 \tab 58 \tab 2 \tab 71 \tab - \tab - \tab - \tab 9 \tab 4 \tab - \tab - \tab 2 \tab 9 \tab - \tab 1 \tab 26 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 8 \tab 4 \tab 1 \tab 1 \tab - \tab 21 \tab 21 \tab 6 \tab - \tab 2 \tab 16 \tab 20 \tab - \tab 14 \tab 4 \tab 1 \tab 6 \tab 4 \tab 19 \tab 2 \tab 3 \tab 1 \tab 3 \tab 4 \tab 4 \tab 6 \tab 102 \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 1 \tab 10 \tab 16 \tab 12 \tab 51 \tab 1 \tab - \tab 8 \tab - \tab - \tab - \tab - \tab 4 \tab 4 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 6 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 6 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 691 \tab 4 \tab 77 \tab 19 \tab 1 \tab - \tab 4 \tab 3 \tab 23 \tab 5 \tab 2 \tab 8 \tab - \tab 23 \tab 12 \tab - \tab - \tab 17 \tab 19 \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab 9 \tab 4 \tab 2 \tab 4 \tab 7 \tab - \tab 5 \tab 3 \tab 3 \tab 2 \tab 1 \tab 1 \tab 7 \tab - \tab 5 \tab 3 \tab 14 \tab 4 \tab 2 \tab 8 \tab - \tab 5 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab - \tab 1 \tab 46 \tab 1 \tab 3 \tab 127 \tab 28 \tab 17 \tab 2 \tab 1 \tab 4 \tab 10 \tab 1 \tab 11 \tab 11 \tab 63 \tab - \tab 2 \tab 3 \tab 2 \tab 3 \tab 9 \tab - \tab 2 \tab 2 \tab 10 \tab 3 \tab 3 \tab - \tab 1 \tab 11 \tab 26 \tab 5 \tab 49 \tab 16 \tab 70 \tab 1 \tab 4 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 3 \tab 12 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 22 \tab 6 \tab 37 \tab - \tab 24 \tab 1 \tab 3 \tab 16 \tab 2 \tab 4 \tab - \tab 2 \tab - \tab 6 \tab 8 \tab 17 \tab 7 \tab 1 \tab 4 \tab - \tab 4 \tab - \tab 2 \tab 3 \tab 2 \tab - \tab 10 \tab 4 \tab 44 \tab 2 \tab 47 \tab - \tab - \tab - \tab 9 \tab 1 \tab - \tab - \tab 6 \tab 8 \tab - \tab 3 \tab 30 \tab 1 \tab 4 \tab 3 \tab 2 \tab - \tab 6 \tab 3 \tab 1 \tab 1 \tab 2 \tab 12 \tab 14 \tab 8 \tab - \tab - \tab 12 \tab 32 \tab - \tab 18 \tab 10 \tab - \tab 8 \tab 1 \tab 16 \tab 1 \tab 6 \tab 2 \tab 5 \tab 4 \tab 4 \tab 3 \tab 65 \tab - \tab 2 \tab - \tab - \tab 7 \tab - \tab 6 \tab 17 \tab 3 \tab 52 \tab 2 \tab 2 \tab 5 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 7 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 47277 \tab 50 \tab 951 \tab 204 \tab 10 \tab 14 \tab 39 \tab 63 \tab 273 \tab 24 \tab 16 \tab 109 \tab 6 \tab 238 \tab 189 \tab 11 \tab 11 \tab 257 \tab 214 \tab 23 \tab 61 \tab 25 \tab 16 \tab 6 \tab 52 \tab 78 \tab 89 \tab 57 \tab 50 \tab 3 \tab 29 \tab 29 \tab 79 \tab 14 \tab 3 \tab 15 \tab 335 \tab 17 \tab 50 \tab 52 \tab 261 \tab 20 \tab 22 \tab 81 \tab 1 \tab 121 \tab 24 \tab 19 \tab 3 \tab 41 \tab 28 \tab 16 \tab 7 \tab 575 \tab 16 \tab 65 \tab 1700 \tab 481 \tab 124 \tab 31 \tab 11 \tab 53 \tab 45 \tab 26 \tab 85 \tab 112 \tab 804 \tab 7 \tab 24 \tab 61 \tab 34 \tab 31 \tab 87 \tab 23 \tab 9 \tab 31 \tab 204 \tab 153 \tab 10 \tab 10 \tab 7 \tab 138 \tab 389 \tab 33 \tab 603 \tab 166 \tab 1189 \tab 44 \tab 109 \tab 7 \tab 15 \tab 4 \tab 19 \tab 38 \tab 51 \tab 131 \tab 2 \tab 9 \tab 50 \tab 21 \tab 16 \tab 9 \tab 21 \tab 127 \tab 147 \tab 595 \tab 60 \tab 621 \tab 3 \tab 44 \tab 188 \tab 40 \tab 35 \tab 10 \tab 19 \tab 18 \tab 59 \tab 121 \tab 107 \tab 62 \tab 8 \tab 13 \tab 1 \tab 31 \tab 3 \tab 9 \tab 84 \tab 17 \tab 10 \tab 95 \tab 66 \tab 449 \tab 19 \tab 834 \tab 4 \tab 5 \tab 2 \tab 99 \tab 16 \tab 1 \tab 2 \tab 32 \tab 56 \tab 4 \tab 18 \tab 298 \tab 16 \tab 30 \tab 11 \tab 7 \tab 2 \tab 59 \tab 26 \tab 15 \tab 15 \tab 15 \tab 174 \tab 216 \tab 112 \tab 8 \tab 20 \tab 127 \tab 235 \tab 3 \tab 95 \tab 62 \tab 14 \tab 55 \tab 31 \tab 165 \tab 17 \tab 31 \tab 14 \tab 49 \tab 30 \tab 32 \tab 32 \tab 720 \tab 3 \tab 6 \tab 2 \tab 16 \tab 84 \tab 5 \tab 128 \tab 138 \tab 62 \tab 529 \tab 28 \tab 15 \tab 49 \tab 3 \tab 3 \tab 6 \tab 7 \tab 40 \tab 10 \tab 4 \tab 9 \tab 9 \tab 4 \tab 17 \tab 3 \tab 18 \tab 11 \tab 9 \tab 16 \tab 8 \tab 4 \tab 1 \tab 10 \tab 25 \tab 42 \tab 8 \tab 3 \tab 11 \tab 1 \tab 8 \tab 18 \tab 14 \tab 16 \tab 14 \tab 4 \tab 8 \tab 3 \tab 3 \tab 9 \tab 14 \tab 3 \tab 5 \tab 8 \tab 1 \tab 5 \tab 4 \tab 1 \tab 43 \tab 3 \tab 6 \tab 4 \tab 16 \tab 9 \tab 4 \tab 5 \tab 6 \tab 6 \tab 4 \tab 3 \tab 1 \tab 9 \tab 13 \tab 1 \tab 2 \tab 2 \tab 3 \tab 6 \tab 4 \tab 6 \tab 3 \tab 3 \tab 11 \tab 5 \tab 3 \tab 9 \tab 6 \tab 1 \tab 5 \tab 1 \tab 6 \tab 5 \tab 8 \tab 3 \tab 3 \tab 12 \tab 3 \tab 6 \tab 3 \tab 8 \tab 3 \tab 11 \tab 1 \tab 2 \tab 4 \tab 2 \tab 2 \tab 2 \tab 2 \tab 7 \tab 5 \tab 5 \tab 4 \tab 3 \tab 2 \tab 7 \tab 2 \tab 1 \tab 2 \tab 4 \tab 7 \tab 1 \tab 1 \tab 24 \tab 1 \tab 6 \tab 2 \tab 9 \tab 7 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 3 \tab 6 \tab 1 \tab 1 \tab 4 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} i don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
