% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_39.R
\name{mastersp}
\alias{mastersp}
\title{Contrast between master and spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
mastersp
}
\details{
Question 114. There are many different ways of picturing God. We'd like to know the kinds of images you are most likely to associate with God. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of God between the two contrasting images? b. 1. Master 7. Spouse
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab don't know \tab master \tab no answer \tab spouse \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 484 \tab 113 \tab 72 \tab 180 \tab 26 \tab 16 \tab 1 \tab 481 \tab 66 \tab 34 \tab 1473 \cr
   1985 \tab - \tab 128 \tab 133 \tab 279 \tab 69 \tab 41 \tab 27 \tab 721 \tab 61 \tab 75 \tab 1534 \cr
   1986 \tab - \tab 137 \tab 109 \tab 257 \tab 43 \tab 31 \tab 16 \tab 770 \tab 57 \tab 50 \tab 1470 \cr
   1987 \tab - \tab 164 \tab 103 \tab 361 \tab 46 \tab 36 \tab 26 \tab 933 \tab 87 \tab 63 \tab 1819 \cr
   1988 \tab - \tab 160 \tab 106 \tab 289 \tab 49 \tab 26 \tab 33 \tab 710 \tab 66 \tab 42 \tab 1481 \cr
   1989 \tab 531 \tab 106 \tab 80 \tab 215 \tab 26 \tab 16 \tab 7 \tab 486 \tab 44 \tab 26 \tab 1537 \cr
   1990 \tab 444 \tab 111 \tab 71 \tab 188 \tab 35 \tab 13 \tab 12 \tab 425 \tab 43 \tab 30 \tab 1372 \cr
   1991 \tab 158 \tab 153 \tab 122 \tab 296 \tab 61 \tab 49 \tab - \tab 434 \tab 205 \tab 39 \tab 1517 \cr
   1993 \tab 526 \tab 110 \tab 98 \tab 200 \tab 35 \tab 25 \tab 8 \tab 489 \tab 63 \tab 52 \tab 1606 \cr
   1994 \tab 1015 \tab 206 \tab 148 \tab 369 \tab 64 \tab 42 \tab 35 \tab 950 \tab 85 \tab 78 \tab 2992 \cr
   1996 \tab 1913 \tab 95 \tab 87 \tab 213 \tab 30 \tab 17 \tab 14 \tab 456 \tab 40 \tab 39 \tab 2904 \cr
   1998 \tab 1387 \tab 180 \tab 116 \tab 313 \tab 37 \tab 34 \tab 25 \tab 612 \tab 64 \tab 64 \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 658 \tab 110 \tab 136 \tab 320 \tab 51 \tab 29 \tab 111 \tab 532 \tab 31 \tab 45 \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 54946 \tab 1773 \tab 1381 \tab 3480 \tab 572 \tab 375 \tab 315 \tab 7999 \tab 912 \tab 637 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} master
\item \code{7} spouse
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
