% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_14.R
\name{mawork}
\alias{mawork}
\title{Mothers employment since marriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
mawork
}
\details{
Question 28. Did your mother ever work for pay for as long as a year, after she was married?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 158 \tab 22 \tab 714 \tab 19 \tab 591 \tab 1504 \cr
   1974 \tab 150 \tab 21 \tab 672 \tab 39 \tab 602 \tab 1484 \cr
   1975 \tab 138 \tab 19 \tab 637 \tab 13 \tab 683 \tab 1490 \cr
   1976 \tab 143 \tab 20 \tab 639 \tab 18 \tab 679 \tab 1499 \cr
   1977 \tab 143 \tab 22 \tab 712 \tab 27 \tab 626 \tab 1530 \cr
   1978 \tab 142 \tab 20 \tab 614 \tab 5 \tab 751 \tab 1532 \cr
   1980 \tab 149 \tab 17 \tab 577 \tab 15 \tab 710 \tab 1468 \cr
   1982 \tab 206 \tab 38 \tab 700 \tab 65 \tab 851 \tab 1860 \cr
   1983 \tab 142 \tab 20 \tab 569 \tab 3 \tab 865 \tab 1599 \cr
   1984 \tab 148 \tab 20 \tab 494 \tab 9 \tab 802 \tab 1473 \cr
   1985 \tab 171 \tab 10 \tab 545 \tab 2 \tab 806 \tab 1534 \cr
   1986 \tab 136 \tab 22 \tab 483 \tab - \tab 829 \tab 1470 \cr
   1987 \tab 202 \tab 25 \tab 567 \tab 7 \tab 1018 \tab 1819 \cr
   1988 \tab 146 \tab 29 \tab 462 \tab 2 \tab 842 \tab 1481 \cr
   1989 \tab 151 \tab 20 \tab 518 \tab 6 \tab 842 \tab 1537 \cr
   1990 \tab 135 \tab 24 \tab 412 \tab 4 \tab 797 \tab 1372 \cr
   1991 \tab 171 \tab 25 \tab 448 \tab 7 \tab 866 \tab 1517 \cr
   1993 \tab 166 \tab 37 \tab 449 \tab 14 \tab 940 \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 47412 \tab 411 \tab 10212 \tab 255 \tab 14100 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
