% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{mawrkslf}
\alias{mawrkslf}
\title{Mother self-emp. or worked for somebody}
\source{
General Social Survey https://gss.norc.org
}
\description{
mawrkslf
}
\details{
Question 10e. Was she self-employed, or did she work for someone else?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab don't know \tab no answer \tab self-employed \tab someone else \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1297 \tab 1 \tab 38 \tab 211 \tab 1445 \tab 2992 \cr
   1996 \tab 1124 \tab 1 \tab 43 \tab 205 \tab 1531 \tab 2904 \cr
   1998 \tab 1182 \tab 2 \tab 27 \tab 174 \tab 1447 \tab 2832 \cr
   2000 \tab 1150 \tab 6 \tab 23 \tab 214 \tab 1424 \tab 2817 \cr
   2002 \tab 1103 \tab - \tab 16 \tab 181 \tab 1465 \tab 2765 \cr
   2004 \tab 1076 \tab 5 \tab 2 \tab 203 \tab 1526 \tab 2812 \cr
   2006 \tab 2637 \tab 2 \tab 30 \tab 206 \tab 1635 \tab 4510 \cr
   2008 \tab 710 \tab 5 \tab 2 \tab 140 \tab 1166 \tab 2023 \cr
   2010 \tab 663 \tab 3 \tab - \tab 166 \tab 1212 \tab 2044 \cr
   2012 \tab 602 \tab 2 \tab 1 \tab 159 \tab 1210 \tab 1974 \cr
   2014 \tab 848 \tab 4 \tab - \tab 190 \tab 1496 \tab 2538 \cr
   2016 \tab 819 \tab 5 \tab 5 \tab 225 \tab 1813 \tab 2867 \cr
   2018 \tab 672 \tab 3 \tab 2 \tab 171 \tab 1500 \tab 2348 \cr
   2021 \tab 1233 \tab - \tab 2 \tab 287 \tab 2510 \tab 4032 \cr
   2022 \tab 966 \tab - \tab 1 \tab 261 \tab 2316 \tab 3544 \cr
   Total \tab 45470 \tab 39 \tab 192 \tab 2993 \tab 23696 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} self-employed
\item \code{2} someone else
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
