% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_106.R
\name{memchurh}
\alias{memchurh}
\title{Membership in church group}
\source{
General Social Survey https://gss.norc.org
}
\description{
memchurh
}
\details{
Question 328. Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type? o. Church-affiliated groups
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab don't know \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 854 \tab 9 \tab 621 \tab - \tab 1484 \cr
   1975 \tab - \tab 877 \tab 23 \tab 588 \tab 2 \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 931 \tab 12 \tab 587 \tab - \tab 1530 \cr
   1978 \tab - \tab 972 \tab 8 \tab 552 \tab - \tab 1532 \cr
   1980 \tab - \tab 1006 \tab 20 \tab 441 \tab 1 \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 992 \tab 7 \tab 600 \tab - \tab 1599 \cr
   1984 \tab - \tab 960 \tab 18 \tab 491 \tab 4 \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 877 \tab 8 \tab 585 \tab - \tab 1470 \cr
   1987 \tab - \tab 1226 \tab 19 \tab 573 \tab 1 \tab 1819 \cr
   1988 \tab 484 \tab 647 \tab 8 \tab 342 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 674 \tab 33 \tab 327 \tab 1 \tab 1537 \cr
   1990 \tab 473 \tab 598 \tab 3 \tab 298 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 670 \tab 9 \tab 338 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 680 \tab 5 \tab 372 \tab - \tab 1606 \cr
   1994 \tab 2481 \tab 339 \tab 2 \tab 170 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1340 \tab 1012 \tab 5 \tab 455 \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 51537 \tab 13315 \tab 189 \tab 7340 \tab 9 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
