% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_106.R
\name{memnat}
\alias{memnat}
\title{memnat}
\source{
General Social Survey https://gss.norc.org
}
\description{
Membership in nationality group
}
\details{
Question 328. Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type? k. Nationality groups
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 2 \tab 1410 \tab 20 \tab 52 \tab 1484 \cr
   1975 \tab - \tab 2 \tab 1417 \tab 34 \tab 37 \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab - \tab 1466 \tab 12 \tab 52 \tab 1530 \cr
   1978 \tab - \tab - \tab 1475 \tab 18 \tab 39 \tab 1532 \cr
   1980 \tab - \tab 2 \tab 1400 \tab 29 \tab 37 \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab - \tab 1535 \tab 7 \tab 57 \tab 1599 \cr
   1984 \tab - \tab 2 \tab 1401 \tab 21 \tab 49 \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab 1392 \tab 10 \tab 68 \tab 1470 \cr
   1987 \tab - \tab 1 \tab 1750 \tab 23 \tab 45 \tab 1819 \cr
   1988 \tab 484 \tab - \tab 963 \tab 12 \tab 22 \tab 1481 \cr
   1989 \tab 502 \tab - \tab 968 \tab 32 \tab 35 \tab 1537 \cr
   1990 \tab 473 \tab 1 \tab 849 \tab 8 \tab 41 \tab 1372 \cr
   1991 \tab 500 \tab 2 \tab 959 \tab 10 \tab 46 \tab 1517 \cr
   1993 \tab 549 \tab 3 \tab 998 \tab 6 \tab 50 \tab 1606 \cr
   1994 \tab 2481 \tab - \tab 491 \tab 2 \tab 18 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1340 \tab - \tab 1428 \tab 5 \tab 39 \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 51537 \tab 15 \tab 19902 \tab 249 \tab 687 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
