% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_106.R
\name{memnum}
\alias{memnum}
\title{memnum}
\source{
General Social Survey https://gss.norc.org
}
\description{
Number of memberships
}
\details{
Question 328. Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type? q. Any membership at all?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 15 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab 14 \tab 16 \tab don't know \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 375 \tab 381 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 281 \tab 185 \tab 114 \tab 67 \tab 28 \tab 21 \tab 11 \tab 8 \tab 3 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 410 \tab 424 \tab 4 \tab - \tab - \tab - \tab - \tab 261 \tab 169 \tab 97 \tab 46 \tab 38 \tab 23 \tab 8 \tab 6 \tab 3 \tab 1 \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 433 \tab 423 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 247 \tab 168 \tab 106 \tab 68 \tab 33 \tab 24 \tab 12 \tab 9 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 426 \tab 418 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 271 \tab 202 \tab 106 \tab 46 \tab 28 \tab 19 \tab 4 \tab 6 \tab 2 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 498 \tab 371 \tab 3 \tab - \tab - \tab - \tab - \tab 242 \tab 164 \tab 81 \tab 43 \tab 30 \tab 13 \tab 9 \tab 5 \tab 8 \tab - \tab 1 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 435 \tab 409 \tab 1 \tab - \tab 2 \tab - \tab - \tab 300 \tab 182 \tab 119 \tab 58 \tab 52 \tab 18 \tab 17 \tab 5 \tab 1 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 474 \tab 347 \tab 8 \tab 3 \tab - \tab - \tab - \tab 256 \tab 158 \tab 89 \tab 53 \tab 36 \tab 19 \tab 12 \tab 6 \tab 11 \tab 1 \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 408 \tab 363 \tab - \tab 1 \tab - \tab - \tab - \tab 259 \tab 168 \tab 127 \tab 66 \tab 37 \tab 24 \tab 8 \tab 6 \tab 3 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 606 \tab 478 \tab 3 \tab - \tab 1 \tab - \tab - \tab 314 \tab 159 \tab 97 \tab 72 \tab 33 \tab 25 \tab 13 \tab 5 \tab 11 \tab - \tab 2 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 294 \tab 256 \tab - \tab - \tab - \tab - \tab - \tab 187 \tab 102 \tab 71 \tab 36 \tab 28 \tab 11 \tab 7 \tab 5 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 502 \tab 313 \tab 243 \tab - \tab - \tab - \tab - \tab - \tab 191 \tab 120 \tab 72 \tab 41 \tab 20 \tab 16 \tab 9 \tab 5 \tab 4 \tab - \tab 1 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 286 \tab 218 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 154 \tab 86 \tab 69 \tab 39 \tab 25 \tab 8 \tab 5 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 500 \tab 329 \tab 242 \tab 2 \tab - \tab 1 \tab - \tab - \tab 182 \tab 107 \tab 73 \tab 37 \tab 22 \tab 10 \tab 7 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 549 \tab 298 \tab 255 \tab 1 \tab 1 \tab - \tab - \tab - \tab 183 \tab 124 \tab 86 \tab 50 \tab 25 \tab 15 \tab 10 \tab 5 \tab 4 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2481 \tab 146 \tab 125 \tab - \tab - \tab - \tab - \tab - \tab 82 \tab 67 \tab 34 \tab 27 \tab 12 \tab 9 \tab 6 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1340 \tab 555 \tab 329 \tab 1 \tab - \tab - \tab 1 \tab - \tab 209 \tab 141 \tab 102 \tab 52 \tab 41 \tab 22 \tab 11 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 51537 \tab 6286 \tab 5282 \tab 32 \tab 12 \tab 10 \tab 4 \tab 1 \tab 3619 \tab 2302 \tab 1443 \tab 801 \tab 488 \tab 277 \tab 149 \tab 78 \tab 62 \tab 2 \tab 4 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
