% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_483.R
\name{meovrwrk}
\alias{meovrwrk}
\title{meovrwrk}
\source{
General Social Survey https://gss.norc.org
}
\description{
Men hurt family when focus on work too much
}
\details{
Question 1297. And, do you agree or disagree: c. Family life often suffers because men concentrate too much on their work.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab agree \tab can't choose \tab disagree \tab neither agree nor disagree \tab no answer \tab strongly agree \tab strongly disagree \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1545 \tab 695 \tab 33 \tab 243 \tab 286 \tab 27 \tab 122 \tab 41 \tab - \tab 2992 \cr
   1996 \tab 1444 \tab 825 \tab 16 \tab 198 \tab 169 \tab 1 \tab 230 \tab 21 \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 940 \tab 877 \tab 43 \tab 361 \tab 331 \tab 22 \tab 209 \tab 34 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 415 \tab 6 \tab 264 \tab 108 \tab - \tab 99 \tab 16 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 460 \tab 4 \tab 188 \tab 135 \tab - \tab 94 \tab 25 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 945 \tab 14 \tab 477 \tab 304 \tab 1 \tab 208 \tab 43 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 653 \tab 12 \tab 310 \tab 161 \tab - \tab 143 \tab 50 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 662 \tab 6 \tab 388 \tab 192 \tab 3 \tab 122 \tab 57 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 558 \tab 11 \tab 382 \tab 170 \tab - \tab 130 \tab 51 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 702 \tab 7 \tab 479 \tab 234 \tab 1 \tab 176 \tab 76 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 819 \tab 9 \tab 536 \tab 257 \tab - \tab 171 \tab 96 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 644 \tab 11 \tab 475 \tab 220 \tab 2 \tab 134 \tab 73 \tab - \tab 2348 \cr
   2021 \tab 1315 \tab 886 \tab 1 \tab 487 \tab 1001 \tab - \tab 202 \tab 138 \tab 2 \tab 4032 \cr
   2022 \tab 1168 \tab 885 \tab 15 \tab 537 \tab 618 \tab 1 \tab 201 \tab 117 \tab 2 \tab 3544 \cr
   Total \tab 49524 \tab 10026 \tab 188 \tab 5325 \tab 4186 \tab 58 \tab 2241 \tab 838 \tab 4 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
