% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_24.R
\name{nataid}
\alias{nataid}
\title{nataid}
\source{
General Social Survey https://gss.norc.org
}
\description{
Foreign aid
}
\details{
Question 68. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. j. Foreign aid.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 302 \tab 82 \tab 1 \tab 63 \tab 1056 \tab - \tab 1504 \cr
   1974 \tab - \tab 258 \tab 59 \tab 3 \tab 45 \tab 1119 \tab - \tab 1484 \cr
   1975 \tab - \tab 247 \tab 73 \tab 1 \tab 81 \tab 1088 \tab - \tab 1490 \cr
   1976 \tab - \tab 268 \tab 56 \tab 5 \tab 44 \tab 1126 \tab - \tab 1499 \cr
   1977 \tab - \tab 359 \tab 106 \tab 3 \tab 52 \tab 1010 \tab - \tab 1530 \cr
   1978 \tab - \tab 366 \tab 88 \tab - \tab 59 \tab 1019 \tab - \tab 1532 \cr
   1980 \tab - \tab 290 \tab 77 \tab 2 \tab 75 \tab 1024 \tab - \tab 1468 \cr
   1982 \tab - \tab 326 \tab 95 \tab 8 \tab 118 \tab 1313 \tab - \tab 1860 \cr
   1983 \tab - \tab 272 \tab 86 \tab 4 \tab 62 \tab 1175 \tab - \tab 1599 \cr
   1984 \tab 983 \tab 100 \tab 24 \tab 4 \tab 21 \tab 341 \tab - \tab 1473 \cr
   1985 \tab 783 \tab 182 \tab 27 \tab 6 \tab 49 \tab 487 \tab - \tab 1534 \cr
   1986 \tab 740 \tab 140 \tab 28 \tab 4 \tab 45 \tab 513 \tab - \tab 1470 \cr
   1987 \tab 1216 \tab 116 \tab 25 \tab 5 \tab 48 \tab 409 \tab - \tab 1819 \cr
   1988 \tab 763 \tab 158 \tab 36 \tab 3 \tab 36 \tab 485 \tab - \tab 1481 \cr
   1989 \tab 769 \tab 164 \tab 47 \tab 6 \tab 34 \tab 517 \tab - \tab 1537 \cr
   1990 \tab 698 \tab 158 \tab 31 \tab 3 \tab 36 \tab 446 \tab - \tab 1372 \cr
   1991 \tab 761 \tab 143 \tab 30 \tab 9 \tab 30 \tab 544 \tab - \tab 1517 \cr
   1993 \tab 807 \tab 165 \tab 36 \tab 4 \tab 39 \tab 555 \tab - \tab 1606 \cr
   1994 \tab 1484 \tab 303 \tab 59 \tab 5 \tab 64 \tab 1077 \tab - \tab 2992 \cr
   1996 \tab 1463 \tab 309 \tab 85 \tab 3 \tab 60 \tab 984 \tab - \tab 2904 \cr
   1998 \tab 1451 \tab 372 \tab 81 \tab 3 \tab 92 \tab 833 \tab - \tab 2832 \cr
   2000 \tab 1408 \tab 398 \tab 94 \tab 4 \tab 114 \tab 799 \tab - \tab 2817 \cr
   2002 \tab 1407 \tab 365 \tab 42 \tab 3 \tab 88 \tab 860 \tab - \tab 2765 \cr
   2004 \tab 1411 \tab 344 \tab 51 \tab 2 \tab 132 \tab 872 \tab - \tab 2812 \cr
   2006 \tab 3026 \tab 358 \tab 54 \tab 5 \tab 163 \tab 904 \tab - \tab 4510 \cr
   2008 \tab 1019 \tab 261 \tab 68 \tab 2 \tab 117 \tab 556 \tab - \tab 2023 \cr
   2010 \tab 1030 \tab 305 \tab 39 \tab 2 \tab 88 \tab 580 \tab - \tab 2044 \cr
   2012 \tab 976 \tab 257 \tab 65 \tab 3 \tab 79 \tab 594 \tab - \tab 1974 \cr
   2014 \tab 1264 \tab 345 \tab 64 \tab 1 \tab 89 \tab 775 \tab - \tab 2538 \cr
   2016 \tab 1430 \tab 452 \tab 86 \tab 6 \tab 149 \tab 744 \tab - \tab 2867 \cr
   2018 \tab 1160 \tab 394 \tab 97 \tab 2 \tab 183 \tab 512 \tab - \tab 2348 \cr
   2021 \tab 2045 \tab 686 \tab 6 \tab - \tab 177 \tab 1106 \tab 12 \tab 4032 \cr
   2022 \tab 1764 \tab 514 \tab 70 \tab 3 \tab 209 \tab 979 \tab 5 \tab 3544 \cr
   Total \tab 31471 \tab 9677 \tab 1967 \tab 115 \tab 2741 \tab 26402 \tab 17 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
