% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_26.R
\name{natarmsy}
\alias{natarmsy}
\title{natarmsy}
\source{
General Social Survey https://gss.norc.org
}
\description{
National defense:version y
}
\details{
Question 69. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. i. National defense.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 974 \tab 192 \tab 23 \tab 12 \tab 87 \tab 185 \tab - \tab 1473 \cr
   1985 \tab 751 \tab 302 \tab 36 \tab 2 \tab 123 \tab 320 \tab - \tab 1534 \cr
   1986 \tab 730 \tab 332 \tab 29 \tab 6 \tab 119 \tab 254 \tab - \tab 1470 \cr
   1987 \tab 603 \tab 475 \tab 63 \tab 13 \tab 211 \tab 454 \tab - \tab 1819 \cr
   1988 \tab 718 \tab 291 \tab 51 \tab 10 \tab 134 \tab 277 \tab - \tab 1481 \cr
   1989 \tab 768 \tab 312 \tab 45 \tab 9 \tab 116 \tab 287 \tab - \tab 1537 \cr
   1990 \tab 674 \tab 294 \tab 27 \tab 15 \tab 65 \tab 297 \tab - \tab 1372 \cr
   1991 \tab 756 \tab 425 \tab 30 \tab 15 \tab 96 \tab 195 \tab - \tab 1517 \cr
   1993 \tab 799 \tab 315 \tab 31 \tab 12 \tab 94 \tab 355 \tab - \tab 1606 \cr
   1994 \tab 1508 \tab 676 \tab 73 \tab 38 \tab 233 \tab 464 \tab - \tab 2992 \cr
   1996 \tab 1441 \tab 648 \tab 86 \tab 60 \tab 231 \tab 438 \tab - \tab 2904 \cr
   1998 \tab 1381 \tab 667 \tab 85 \tab 9 \tab 254 \tab 436 \tab - \tab 2832 \cr
   2000 \tab 1409 \tab 618 \tab 87 \tab 4 \tab 351 \tab 348 \tab - \tab 2817 \cr
   2002 \tab 1358 \tab 591 \tab 57 \tab 2 \tab 477 \tab 280 \tab - \tab 2765 \cr
   2004 \tab 1401 \tab 538 \tab 37 \tab 3 \tab 445 \tab 388 \tab - \tab 2812 \cr
   2006 \tab 3002 \tab 478 \tab 61 \tab 2 \tab 413 \tab 554 \tab - \tab 4510 \cr
   2008 \tab 1004 \tab 345 \tab 36 \tab 3 \tab 268 \tab 367 \tab - \tab 2023 \cr
   2010 \tab 1014 \tab 408 \tab 49 \tab - \tab 243 \tab 330 \tab - \tab 2044 \cr
   2012 \tab 998 \tab 395 \tab 41 \tab 1 \tab 241 \tab 298 \tab - \tab 1974 \cr
   2014 \tab 1274 \tab 436 \tab 53 \tab 2 \tab 392 \tab 381 \tab - \tab 2538 \cr
   2016 \tab 1437 \tab 487 \tab 53 \tab - \tab 525 \tab 365 \tab - \tab 2867 \cr
   2018 \tab 1188 \tab 449 \tab 42 \tab 3 \tab 341 \tab 325 \tab - \tab 2348 \cr
   2021 \tab 1987 \tab 915 \tab 5 \tab - \tab 512 \tab 606 \tab 7 \tab 4032 \cr
   2022 \tab 1780 \tab 684 \tab 50 \tab 1 \tab 550 \tab 476 \tab 3 \tab 3544 \cr
   Total \tab 44534 \tab 11273 \tab 1150 \tab 222 \tab 6521 \tab 8680 \tab 10 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
