% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_25.R
\name{natchld}
\alias{natchld}
\title{Assistance for childcare}
\source{
General Social Survey https://gss.norc.org
}
\description{
natchld
}
\details{
Question 68. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. p. Assistance for childcare.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 758 \tab 235 \tab 5 \tab 1678 \tab 141 \tab - \tab 2817 \cr
   2002 \tab - \tab 861 \tab 166 \tab 6 \tab 1533 \tab 199 \tab - \tab 2765 \cr
   2004 \tab - \tab 933 \tab 179 \tab 2 \tab 1518 \tab 180 \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 1002 \tab 230 \tab 4 \tab 1544 \tab 212 \tab - \tab 4510 \cr
   2008 \tab - \tab 722 \tab 151 \tab 1 \tab 1024 \tab 125 \tab - \tab 2023 \cr
   2010 \tab - \tab 764 \tab 147 \tab 3 \tab 982 \tab 148 \tab - \tab 2044 \cr
   2012 \tab - \tab 756 \tab 150 \tab 6 \tab 890 \tab 172 \tab - \tab 1974 \cr
   2014 \tab - \tab 936 \tab 186 \tab 1 \tab 1203 \tab 212 \tab - \tab 2538 \cr
   2016 \tab - \tab 981 \tab 161 \tab 6 \tab 1556 \tab 163 \tab - \tab 2867 \cr
   2018 \tab - \tab 747 \tab 133 \tab - \tab 1345 \tab 123 \tab - \tab 2348 \cr
   2021 \tab - \tab 1418 \tab 17 \tab 2 \tab 2244 \tab 322 \tab 29 \tab 4032 \cr
   2022 \tab - \tab 933 \tab 120 \tab - \tab 2240 \tab 240 \tab 11 \tab 3544 \cr
   Total \tab 39634 \tab 10811 \tab 1875 \tab 36 \tab 17757 \tab 2237 \tab 40 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
