% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_24.R
\name{natcrime}
\alias{natcrime}
\title{Halting rising crime rate}
\source{
General Social Survey https://gss.norc.org
}
\description{
natcrime
}
\details{
Question 68. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. e. Halting the rising crime rate.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 368 \tab 92 \tab 7 \tab 967 \tab 70 \tab - \tab 1504 \cr
   1974 \tab - \tab 346 \tab 76 \tab 3 \tab 987 \tab 72 \tab - \tab 1484 \cr
   1975 \tab - \tab 345 \tab 84 \tab 6 \tab 974 \tab 81 \tab - \tab 1490 \cr
   1976 \tab - \tab 316 \tab 76 \tab 10 \tab 979 \tab 118 \tab - \tab 1499 \cr
   1977 \tab - \tab 339 \tab 93 \tab 6 \tab 1002 \tab 90 \tab - \tab 1530 \cr
   1978 \tab - \tab 385 \tab 66 \tab 6 \tab 982 \tab 93 \tab - \tab 1532 \cr
   1980 \tab - \tab 304 \tab 63 \tab 5 \tab 1008 \tab 88 \tab - \tab 1468 \cr
   1982 \tab - \tab 322 \tab 79 \tab 13 \tab 1341 \tab 105 \tab - \tab 1860 \cr
   1983 \tab - \tab 385 \tab 70 \tab 8 \tab 1060 \tab 76 \tab - \tab 1599 \cr
   1984 \tab 983 \tab 119 \tab 15 \tab 6 \tab 328 \tab 22 \tab - \tab 1473 \cr
   1985 \tab 783 \tab 207 \tab 29 \tab 4 \tab 470 \tab 41 \tab - \tab 1534 \cr
   1986 \tab 740 \tab 196 \tab 32 \tab 5 \tab 463 \tab 34 \tab - \tab 1470 \cr
   1987 \tab 1216 \tab 135 \tab 19 \tab 10 \tab 414 \tab 25 \tab - \tab 1819 \cr
   1988 \tab 763 \tab 165 \tab 33 \tab 6 \tab 486 \tab 28 \tab - \tab 1481 \cr
   1989 \tab 769 \tab 150 \tab 24 \tab 5 \tab 553 \tab 36 \tab - \tab 1537 \cr
   1990 \tab 698 \tab 147 \tab 29 \tab 3 \tab 468 \tab 27 \tab - \tab 1372 \cr
   1991 \tab 761 \tab 203 \tab 25 \tab 8 \tab 484 \tab 36 \tab - \tab 1517 \cr
   1993 \tab 807 \tab 163 \tab 29 \tab 5 \tab 563 \tab 39 \tab - \tab 1606 \cr
   1994 \tab 1484 \tab 241 \tab 39 \tab 4 \tab 1131 \tab 93 \tab - \tab 2992 \cr
   1996 \tab 1463 \tab 330 \tab 45 \tab 4 \tab 965 \tab 97 \tab - \tab 2904 \cr
   1998 \tab 1451 \tab 392 \tab 47 \tab 4 \tab 842 \tab 96 \tab - \tab 2832 \cr
   2000 \tab 1408 \tab 454 \tab 44 \tab 8 \tab 827 \tab 76 \tab - \tab 2817 \cr
   2002 \tab 1407 \tab 473 \tab 35 \tab 2 \tab 758 \tab 90 \tab - \tab 2765 \cr
   2004 \tab 1411 \tab 490 \tab 41 \tab 2 \tab 796 \tab 72 \tab - \tab 2812 \cr
   2006 \tab 3026 \tab 456 \tab 35 \tab 1 \tab 897 \tab 95 \tab - \tab 4510 \cr
   2008 \tab 1019 \tab 296 \tab 39 \tab - \tab 601 \tab 68 \tab - \tab 2023 \cr
   2010 \tab 1030 \tab 328 \tab 32 \tab 2 \tab 581 \tab 71 \tab - \tab 2044 \cr
   2012 \tab 976 \tab 319 \tab 37 \tab 1 \tab 573 \tab 68 \tab - \tab 1974 \cr
   2014 \tab 1264 \tab 371 \tab 41 \tab 1 \tab 755 \tab 106 \tab - \tab 2538 \cr
   2016 \tab 1430 \tab 354 \tab 45 \tab 3 \tab 945 \tab 90 \tab - \tab 2867 \cr
   2018 \tab 1160 \tab 285 \tab 52 \tab 3 \tab 771 \tab 77 \tab - \tab 2348 \cr
   2021 \tab 2045 \tab 556 \tab 8 \tab - \tab 1249 \tab 164 \tab 10 \tab 4032 \cr
   2022 \tab 1764 \tab 342 \tab 38 \tab 1 \tab 1254 \tab 142 \tab 3 \tab 3544 \cr
   Total \tab 31471 \tab 10282 \tab 1512 \tab 152 \tab 26474 \tab 2486 \tab 13 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
