% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_26.R
\name{natcrimy}
\alias{natcrimy}
\title{natcrimy}
\source{
General Social Survey https://gss.norc.org
}
\description{
Law enforcement:version y
}
\details{
Question 69. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. e. Law enforcement.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 974 \tab 173 \tab 13 \tab 12 \tab 268 \tab 33 \tab - \tab 1473 \cr
   1985 \tab 751 \tab 262 \tab 33 \tab 5 \tab 440 \tab 43 \tab - \tab 1534 \cr
   1986 \tab 730 \tab 284 \tab 27 \tab 8 \tab 373 \tab 48 \tab - \tab 1470 \cr
   1987 \tab 603 \tab 451 \tab 42 \tab 13 \tab 627 \tab 83 \tab - \tab 1819 \cr
   1988 \tab 718 \tab 273 \tab 28 \tab 11 \tab 409 \tab 42 \tab - \tab 1481 \cr
   1989 \tab 768 \tab 221 \tab 29 \tab 10 \tab 460 \tab 49 \tab - \tab 1537 \cr
   1990 \tab 674 \tab 238 \tab 24 \tab 16 \tab 380 \tab 40 \tab - \tab 1372 \cr
   1991 \tab 756 \tab 271 \tab 22 \tab 14 \tab 407 \tab 47 \tab - \tab 1517 \cr
   1993 \tab 799 \tab 251 \tab 33 \tab 12 \tab 466 \tab 45 \tab - \tab 1606 \cr
   1994 \tab 1508 \tab 399 \tab 37 \tab 38 \tab 909 \tab 101 \tab - \tab 2992 \cr
   1996 \tab 1441 \tab 436 \tab 44 \tab 59 \tab 801 \tab 123 \tab - \tab 2904 \cr
   1998 \tab 1381 \tab 511 \tab 42 \tab 9 \tab 776 \tab 113 \tab - \tab 2832 \cr
   2000 \tab 1409 \tab 534 \tab 51 \tab 5 \tab 706 \tab 112 \tab - \tab 2817 \cr
   2002 \tab 1358 \tab 568 \tab 30 \tab 4 \tab 663 \tab 142 \tab - \tab 2765 \cr
   2004 \tab 1401 \tab 506 \tab 33 \tab 1 \tab 738 \tab 133 \tab - \tab 2812 \cr
   2006 \tab 3002 \tab 529 \tab 35 \tab - \tab 816 \tab 128 \tab - \tab 4510 \cr
   2008 \tab 1005 \tab 365 \tab 17 \tab 1 \tab 543 \tab 92 \tab - \tab 2023 \cr
   2010 \tab 1014 \tab 389 \tab 24 \tab 1 \tab 492 \tab 124 \tab - \tab 2044 \cr
   2012 \tab 998 \tab 381 \tab 28 \tab 1 \tab 460 \tab 106 \tab - \tab 1974 \cr
   2014 \tab 1274 \tab 484 \tab 29 \tab 1 \tab 581 \tab 169 \tab - \tab 2538 \cr
   2016 \tab 1437 \tab 491 \tab 21 \tab 1 \tab 766 \tab 151 \tab - \tab 2867 \cr
   2018 \tab 1188 \tab 379 \tab 41 \tab 3 \tab 615 \tab 122 \tab - \tab 2348 \cr
   2021 \tab 1987 \tab 727 \tab - \tab - \tab 853 \tab 457 \tab 8 \tab 4032 \cr
   2022 \tab 1780 \tab 489 \tab 32 \tab - \tab 961 \tab 281 \tab 1 \tab 3544 \cr
   Total \tab 44535 \tab 9612 \tab 715 \tab 225 \tab 14510 \tab 2784 \tab 9 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
