% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_24.R
\name{natdrug}
\alias{natdrug}
\title{Dealing with drug addiction}
\source{
General Social Survey https://gss.norc.org
}
\description{
natdrug
}
\details{
Question 68. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. f. Dealing with drug addiction.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 325 \tab 94 \tab 11 \tab 984 \tab 90 \tab - \tab 1504 \cr
   1974 \tab - \tab 413 \tab 82 \tab 6 \tab 887 \tab 96 \tab - \tab 1484 \cr
   1975 \tab - \tab 429 \tab 112 \tab 8 \tab 817 \tab 124 \tab - \tab 1490 \cr
   1976 \tab - \tab 401 \tab 103 \tab 6 \tab 876 \tab 113 \tab - \tab 1499 \cr
   1977 \tab - \tab 441 \tab 110 \tab 10 \tab 839 \tab 130 \tab - \tab 1530 \cr
   1978 \tab - \tab 474 \tab 75 \tab 5 \tab 844 \tab 134 \tab - \tab 1532 \cr
   1980 \tab - \tab 367 \tab 107 \tab 8 \tab 873 \tab 113 \tab - \tab 1468 \cr
   1982 \tab - \tab 473 \tab 134 \tab 9 \tab 1085 \tab 159 \tab - \tab 1860 \cr
   1983 \tab - \tab 474 \tab 84 \tab 8 \tab 947 \tab 86 \tab - \tab 1599 \cr
   1984 \tab 983 \tab 133 \tab 18 \tab 5 \tab 306 \tab 28 \tab - \tab 1473 \cr
   1985 \tab 783 \tab 212 \tab 33 \tab 3 \tab 462 \tab 41 \tab - \tab 1534 \cr
   1986 \tab 740 \tab 229 \tab 31 \tab 4 \tab 422 \tab 44 \tab - \tab 1470 \cr
   1987 \tab 1216 \tab 152 \tab 17 \tab 3 \tab 406 \tab 25 \tab - \tab 1819 \cr
   1988 \tab 763 \tab 174 \tab 22 \tab 5 \tab 488 \tab 29 \tab - \tab 1481 \cr
   1989 \tab 769 \tab 147 \tab 28 \tab 4 \tab 541 \tab 48 \tab - \tab 1537 \cr
   1990 \tab 698 \tab 174 \tab 22 \tab 5 \tab 429 \tab 44 \tab - \tab 1372 \cr
   1991 \tab 761 \tab 237 \tab 26 \tab 7 \tab 431 \tab 55 \tab - \tab 1517 \cr
   1993 \tab 807 \tab 217 \tab 36 \tab 7 \tab 476 \tab 63 \tab - \tab 1606 \cr
   1994 \tab 1484 \tab 397 \tab 77 \tab 5 \tab 899 \tab 130 \tab - \tab 2992 \cr
   1996 \tab 1463 \tab 382 \tab 56 \tab 6 \tab 838 \tab 159 \tab - \tab 2904 \cr
   1998 \tab 1451 \tab 386 \tab 58 \tab 5 \tab 803 \tab 129 \tab - \tab 2832 \cr
   2000 \tab 1408 \tab 390 \tab 74 \tab 7 \tab 821 \tab 117 \tab - \tab 2817 \cr
   2002 \tab 1407 \tab 409 \tab 44 \tab 4 \tab 774 \tab 127 \tab - \tab 2765 \cr
   2004 \tab 1411 \tab 456 \tab 55 \tab 3 \tab 757 \tab 130 \tab - \tab 2812 \cr
   2006 \tab 3026 \tab 401 \tab 54 \tab 2 \tab 906 \tab 121 \tab - \tab 4510 \cr
   2008 \tab 1019 \tab 308 \tab 47 \tab 1 \tab 556 \tab 92 \tab - \tab 2023 \cr
   2010 \tab 1030 \tab 327 \tab 36 \tab 2 \tab 544 \tab 105 \tab - \tab 2044 \cr
   2012 \tab 976 \tab 290 \tab 52 \tab 2 \tab 544 \tab 110 \tab - \tab 1974 \cr
   2014 \tab 1264 \tab 349 \tab 56 \tab - \tab 724 \tab 145 \tab - \tab 2538 \cr
   2016 \tab 1430 \tab 367 \tab 40 \tab 2 \tab 918 \tab 110 \tab - \tab 2867 \cr
   2018 \tab 1160 \tab 245 \tab 36 \tab 2 \tab 808 \tab 97 \tab - \tab 2348 \cr
   2021 \tab 2045 \tab 581 \tab 8 \tab 1 \tab 1216 \tab 167 \tab 14 \tab 4032 \cr
   2022 \tab 1764 \tab 360 \tab 34 \tab 1 \tab 1221 \tab 157 \tab 7 \tab 3544 \cr
   Total \tab 31471 \tab 11120 \tab 1861 \tab 157 \tab 24442 \tab 3318 \tab 21 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
