% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_26.R
\name{nateducy}
\alias{nateducy}
\title{Education:version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
nateducy
}
\details{
Question 69. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. g. Education
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 974 \tab 133 \tab 8 \tab 11 \tab 311 \tab 36 \tab - \tab 1473 \cr
   1985 \tab 751 \tab 202 \tab 15 \tab 2 \tab 510 \tab 54 \tab - \tab 1534 \cr
   1986 \tab 730 \tab 197 \tab 12 \tab 8 \tab 486 \tab 37 \tab - \tab 1470 \cr
   1987 \tab 603 \tab 307 \tab 28 \tab 14 \tab 811 \tab 56 \tab - \tab 1819 \cr
   1988 \tab 718 \tab 191 \tab 12 \tab 8 \tab 517 \tab 35 \tab - \tab 1481 \cr
   1989 \tab 768 \tab 175 \tab 8 \tab 7 \tab 551 \tab 28 \tab - \tab 1537 \cr
   1990 \tab 674 \tab 149 \tab 16 \tab 14 \tab 497 \tab 22 \tab - \tab 1372 \cr
   1991 \tab 756 \tab 175 \tab 17 \tab 15 \tab 523 \tab 31 \tab - \tab 1517 \cr
   1993 \tab 799 \tab 178 \tab 15 \tab 13 \tab 562 \tab 39 \tab - \tab 1606 \cr
   1994 \tab 1508 \tab 313 \tab 16 \tab 36 \tab 1048 \tab 71 \tab - \tab 2992 \cr
   1996 \tab 1441 \tab 262 \tab 20 \tab 59 \tab 1053 \tab 69 \tab - \tab 2904 \cr
   1998 \tab 1381 \tab 274 \tab 25 \tab 8 \tab 1058 \tab 86 \tab - \tab 2832 \cr
   2000 \tab 1409 \tab 273 \tab 17 \tab 3 \tab 1053 \tab 62 \tab - \tab 2817 \cr
   2002 \tab 1358 \tab 252 \tab 15 \tab 2 \tab 1072 \tab 66 \tab - \tab 2765 \cr
   2004 \tab 1401 \tab 239 \tab 10 \tab 1 \tab 1089 \tab 72 \tab - \tab 2812 \cr
   2006 \tab 3002 \tab 281 \tab 21 \tab - \tab 1111 \tab 95 \tab - \tab 4510 \cr
   2008 \tab 1004 \tab 196 \tab 11 \tab - \tab 756 \tab 56 \tab - \tab 2023 \cr
   2010 \tab 1014 \tab 190 \tab 6 \tab 1 \tab 780 \tab 53 \tab - \tab 2044 \cr
   2012 \tab 998 \tab 167 \tab 11 \tab - \tab 727 \tab 71 \tab - \tab 1974 \cr
   2014 \tab 1274 \tab 220 \tab 18 \tab - \tab 932 \tab 94 \tab - \tab 2538 \cr
   2016 \tab 1437 \tab 257 \tab 12 \tab - \tab 1060 \tab 101 \tab - \tab 2867 \cr
   2018 \tab 1188 \tab 181 \tab 9 \tab - \tab 904 \tab 66 \tab - \tab 2348 \cr
   2021 \tab 1987 \tab 346 \tab 4 \tab - \tab 1532 \tab 153 \tab 10 \tab 4032 \cr
   2022 \tab 1780 \tab 277 \tab 7 \tab - \tab 1357 \tab 121 \tab 2 \tab 3544 \cr
   Total \tab 44534 \tab 5435 \tab 333 \tab 202 \tab 20300 \tab 1574 \tab 12 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
