% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_24.R
\name{natenvir}
\alias{natenvir}
\title{Improving & protecting environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
natenvir
}
\details{
Question 68. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. b. Improving and protecting the environment.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 387 \tab 85 \tab 6 \tab 915 \tab 111 \tab - \tab 1504 \cr
   1974 \tab - \tab 393 \tab 98 \tab 8 \tab 871 \tab 114 \tab - \tab 1484 \cr
   1975 \tab - \tab 459 \tab 92 \tab - \tab 795 \tab 144 \tab - \tab 1490 \cr
   1976 \tab - \tab 468 \tab 69 \tab 5 \tab 818 \tab 139 \tab - \tab 1499 \cr
   1977 \tab - \tab 522 \tab 110 \tab 6 \tab 724 \tab 168 \tab - \tab 1530 \cr
   1978 \tab - \tab 501 \tab 80 \tab 4 \tab 801 \tab 146 \tab - \tab 1532 \cr
   1980 \tab - \tab 454 \tab 83 \tab 3 \tab 702 \tab 226 \tab - \tab 1468 \cr
   1982 \tab - \tab 559 \tab 124 \tab 4 \tab 964 \tab 209 \tab - \tab 1860 \cr
   1983 \tab - \tab 498 \tab 98 \tab 7 \tab 865 \tab 131 \tab - \tab 1599 \cr
   1984 \tab 983 \tab 159 \tab 19 \tab 5 \tab 286 \tab 21 \tab - \tab 1473 \cr
   1985 \tab 783 \tab 236 \tab 30 \tab 4 \tab 421 \tab 60 \tab - \tab 1534 \cr
   1986 \tab 740 \tab 213 \tab 43 \tab 5 \tab 431 \tab 38 \tab - \tab 1470 \cr
   1987 \tab 1216 \tab 148 \tab 30 \tab 4 \tab 395 \tab 26 \tab - \tab 1819 \cr
   1988 \tab 763 \tab 188 \tab 29 \tab 4 \tab 464 \tab 33 \tab - \tab 1481 \cr
   1989 \tab 769 \tab 155 \tab 32 \tab 5 \tab 548 \tab 28 \tab - \tab 1537 \cr
   1990 \tab 698 \tab 127 \tab 35 \tab 5 \tab 476 \tab 31 \tab - \tab 1372 \cr
   1991 \tab 761 \tab 180 \tab 23 \tab 5 \tab 511 \tab 37 \tab - \tab 1517 \cr
   1993 \tab 807 \tab 239 \tab 36 \tab 5 \tab 445 \tab 74 \tab - \tab 1606 \cr
   1994 \tab 1484 \tab 443 \tab 51 \tab 2 \tab 885 \tab 127 \tab - \tab 2992 \cr
   1996 \tab 1463 \tab 391 \tab 61 \tab 1 \tab 836 \tab 152 \tab - \tab 2904 \cr
   1998 \tab 1451 \tab 386 \tab 59 \tab 4 \tab 826 \tab 106 \tab - \tab 2832 \cr
   2000 \tab 1408 \tab 385 \tab 52 \tab 2 \tab 858 \tab 112 \tab - \tab 2817 \cr
   2002 \tab 1407 \tab 439 \tab 33 \tab 3 \tab 793 \tab 90 \tab - \tab 2765 \cr
   2004 \tab 1411 \tab 405 \tab 38 \tab 1 \tab 862 \tab 95 \tab - \tab 2812 \cr
   2006 \tab 3026 \tab 365 \tab 38 \tab - \tab 992 \tab 89 \tab - \tab 4510 \cr
   2008 \tab 1019 \tab 236 \tab 28 \tab - \tab 659 \tab 81 \tab - \tab 2023 \cr
   2010 \tab 1030 \tab 297 \tab 36 \tab 2 \tab 557 \tab 122 \tab - \tab 2044 \cr
   2012 \tab 976 \tab 306 \tab 25 \tab 2 \tab 559 \tab 106 \tab - \tab 1974 \cr
   2014 \tab 1264 \tab 394 \tab 29 \tab 1 \tab 731 \tab 119 \tab - \tab 2538 \cr
   2016 \tab 1430 \tab 409 \tab 35 \tab 1 \tab 878 \tab 114 \tab - \tab 2867 \cr
   2018 \tab 1160 \tab 284 \tab 30 \tab 1 \tab 790 \tab 83 \tab - \tab 2348 \cr
   2021 \tab 2045 \tab 413 \tab 6 \tab - \tab 1376 \tab 190 \tab 2 \tab 4032 \cr
   2022 \tab 1764 \tab 349 \tab 27 \tab - \tab 1210 \tab 193 \tab 1 \tab 3544 \cr
   Total \tab 31471 \tab 11388 \tab 1664 \tab 105 \tab 24244 \tab 3515 \tab 3 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
