% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_24.R
\name{natfare}
\alias{natfare}
\title{Welfare}
\source{
General Social Survey https://gss.norc.org
}
\description{
natfare
}
\details{
Question 68. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. k. Welfare.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 365 \tab 65 \tab 7 \tab 297 \tab 770 \tab - \tab 1504 \cr
   1974 \tab - \tab 471 \tab 59 \tab 3 \tab 328 \tab 623 \tab - \tab 1484 \cr
   1975 \tab - \tab 423 \tab 79 \tab 6 \tab 347 \tab 635 \tab - \tab 1490 \cr
   1976 \tab - \tab 335 \tab 64 \tab 6 \tab 199 \tab 895 \tab - \tab 1499 \cr
   1977 \tab - \tab 351 \tab 75 \tab 6 \tab 188 \tab 910 \tab - \tab 1530 \cr
   1978 \tab - \tab 382 \tab 56 \tab 3 \tab 199 \tab 892 \tab - \tab 1532 \cr
   1980 \tab - \tab 378 \tab 62 \tab 5 \tab 196 \tab 827 \tab - \tab 1468 \cr
   1982 \tab - \tab 491 \tab 91 \tab 3 \tab 496 \tab 779 \tab - \tab 1860 \cr
   1983 \tab - \tab 446 \tab 64 \tab 5 \tab 341 \tab 743 \tab - \tab 1599 \cr
   1984 \tab 983 \tab 162 \tab 13 \tab 6 \tab 116 \tab 193 \tab - \tab 1473 \cr
   1985 \tab 783 \tab 246 \tab 30 \tab 2 \tab 139 \tab 334 \tab - \tab 1534 \cr
   1986 \tab 740 \tab 244 \tab 26 \tab 4 \tab 162 \tab 294 \tab - \tab 1470 \cr
   1987 \tab 1216 \tab 176 \tab 30 \tab 5 \tab 156 \tab 236 \tab - \tab 1819 \cr
   1988 \tab 763 \tab 226 \tab 24 \tab 9 \tab 162 \tab 297 \tab - \tab 1481 \cr
   1989 \tab 769 \tab 228 \tab 42 \tab 6 \tab 174 \tab 318 \tab - \tab 1537 \cr
   1990 \tab 698 \tab 232 \tab 35 \tab 4 \tab 148 \tab 255 \tab - \tab 1372 \cr
   1991 \tab 761 \tab 261 \tab 39 \tab 10 \tab 166 \tab 280 \tab - \tab 1517 \cr
   1993 \tab 807 \tab 198 \tab 40 \tab 3 \tab 130 \tab 428 \tab - \tab 1606 \cr
   1994 \tab 1484 \tab 359 \tab 54 \tab 6 \tab 189 \tab 900 \tab - \tab 2992 \cr
   1996 \tab 1463 \tab 373 \tab 45 \tab 2 \tab 217 \tab 804 \tab - \tab 2904 \cr
   1998 \tab 1451 \tab 498 \tab 62 \tab 2 \tab 221 \tab 598 \tab - \tab 2832 \cr
   2000 \tab 1408 \tab 531 \tab 72 \tab 6 \tab 282 \tab 518 \tab - \tab 2817 \cr
   2002 \tab 1407 \tab 502 \tab 41 \tab 3 \tab 279 \tab 533 \tab - \tab 2765 \cr
   2004 \tab 1411 \tab 475 \tab 40 \tab 1 \tab 331 \tab 554 \tab - \tab 2812 \cr
   2006 \tab 3026 \tab 515 \tab 48 \tab 2 \tab 380 \tab 539 \tab - \tab 4510 \cr
   2008 \tab 1019 \tab 349 \tab 48 \tab 1 \tab 251 \tab 355 \tab - \tab 2023 \cr
   2010 \tab 1030 \tab 338 \tab 33 \tab 1 \tab 234 \tab 408 \tab - \tab 2044 \cr
   2012 \tab 976 \tab 333 \tab 38 \tab 3 \tab 197 \tab 427 \tab - \tab 1974 \cr
   2014 \tab 1264 \tab 372 \tab 49 \tab 1 \tab 250 \tab 602 \tab - \tab 2538 \cr
   2016 \tab 1430 \tab 477 \tab 47 \tab 5 \tab 316 \tab 592 \tab - \tab 2867 \cr
   2018 \tab 1160 \tab 407 \tab 54 \tab 1 \tab 285 \tab 441 \tab - \tab 2348 \cr
   2021 \tab 2045 \tab 717 \tab 7 \tab - \tab 604 \tab 649 \tab 10 \tab 4032 \cr
   2022 \tab 1764 \tab 597 \tab 54 \tab 1 \tab 589 \tab 534 \tab 5 \tab 3544 \cr
   Total \tab 31471 \tab 12458 \tab 1586 \tab 128 \tab 8569 \tab 18163 \tab 15 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
