% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_25.R
\name{nathealy}
\alias{nathealy}
\title{nathealy}
\source{
General Social Survey https://gss.norc.org
}
\description{
Health:version y
}
\details{
Question 69. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. c. Health
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 974 \tab 158 \tab 19 \tab 11 \tab 270 \tab 41 \tab - \tab 1473 \cr
   1985 \tab 751 \tab 253 \tab 19 \tab 3 \tab 432 \tab 76 \tab - \tab 1534 \cr
   1986 \tab 730 \tab 226 \tab 23 \tab 10 \tab 431 \tab 50 \tab - \tab 1470 \cr
   1987 \tab 603 \tab 296 \tab 36 \tab 12 \tab 806 \tab 66 \tab - \tab 1819 \cr
   1988 \tab 718 \tab 179 \tab 22 \tab 9 \tab 514 \tab 39 \tab - \tab 1481 \cr
   1989 \tab 768 \tab 171 \tab 12 \tab 9 \tab 536 \tab 41 \tab - \tab 1537 \cr
   1990 \tab 674 \tab 168 \tab 21 \tab 15 \tab 457 \tab 37 \tab - \tab 1372 \cr
   1991 \tab 756 \tab 189 \tab 18 \tab 16 \tab 509 \tab 29 \tab - \tab 1517 \cr
   1993 \tab 799 \tab 99 \tab 32 \tab 15 \tab 555 \tab 106 \tab - \tab 1606 \cr
   1994 \tab 1508 \tab 294 \tab 58 \tab 38 \tab 906 \tab 188 \tab - \tab 2992 \cr
   1996 \tab 1441 \tab 341 \tab 34 \tab 57 \tab 879 \tab 152 \tab - \tab 2904 \cr
   1998 \tab 1381 \tab 305 \tab 51 \tab 11 \tab 975 \tab 109 \tab - \tab 2832 \cr
   2000 \tab 1409 \tab 306 \tab 38 \tab 3 \tab 976 \tab 85 \tab - \tab 2817 \cr
   2002 \tab 1358 \tab 265 \tab 18 \tab 3 \tab 1026 \tab 95 \tab - \tab 2765 \cr
   2004 \tab 1401 \tab 200 \tab 20 \tab 3 \tab 1093 \tab 95 \tab - \tab 2812 \cr
   2006 \tab 3002 \tab 258 \tab 27 \tab - \tab 1086 \tab 137 \tab - \tab 4510 \cr
   2008 \tab 1005 \tab 147 \tab 17 \tab - \tab 740 \tab 114 \tab - \tab 2023 \cr
   2010 \tab 1014 \tab 184 \tab 33 \tab 2 \tab 601 \tab 210 \tab - \tab 2044 \cr
   2012 \tab 998 \tab 207 \tab 36 \tab - \tab 572 \tab 161 \tab - \tab 1974 \cr
   2014 \tab 1274 \tab 272 \tab 34 \tab 3 \tab 701 \tab 254 \tab - \tab 2538 \cr
   2016 \tab 1437 \tab 303 \tab 31 \tab - \tab 887 \tab 209 \tab - \tab 2867 \cr
   2018 \tab 1188 \tab 220 \tab 22 \tab - \tab 802 \tab 116 \tab - \tab 2348 \cr
   2021 \tab 1987 \tab 432 \tab 4 \tab - \tab 1333 \tab 260 \tab 16 \tab 4032 \cr
   2022 \tab 1780 \tab 330 \tab 22 \tab - \tab 1208 \tab 201 \tab 3 \tab 3544 \cr
   Total \tab 44535 \tab 5803 \tab 647 \tab 220 \tab 18295 \tab 2871 \tab 19 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
