% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_25.R
\name{natmass}
\alias{natmass}
\title{Mass transportation}
\source{
General Social Survey https://gss.norc.org
}
\description{
natmass
}
\details{
Question 68. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. n. Mass Transportation.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 484 \tab 445 \tab 91 \tab 17 \tab 326 \tab 110 \tab - \tab 1473 \cr
   1985 \tab - \tab 729 \tab 156 \tab 6 \tab 458 \tab 185 \tab - \tab 1534 \cr
   1986 \tab - \tab 729 \tab 132 \tab 13 \tab 415 \tab 181 \tab - \tab 1470 \cr
   1987 \tab - \tab 835 \tab 191 \tab 19 \tab 544 \tab 230 \tab - \tab 1819 \cr
   1988 \tab - \tab 721 \tab 182 \tab 22 \tab 411 \tab 145 \tab - \tab 1481 \cr
   1989 \tab - \tab 749 \tab 207 \tab 11 \tab 442 \tab 128 \tab - \tab 1537 \cr
   1990 \tab - \tab 629 \tab 156 \tab 20 \tab 455 \tab 112 \tab - \tab 1372 \cr
   1991 \tab - \tab 712 \tab 149 \tab 22 \tab 500 \tab 134 \tab - \tab 1517 \cr
   1993 \tab - \tab 702 \tab 181 \tab 16 \tab 554 \tab 153 \tab - \tab 1606 \cr
   1994 \tab - \tab 1423 \tab 254 \tab 44 \tab 1036 \tab 235 \tab - \tab 2992 \cr
   1996 \tab - \tab 1334 \tab 314 \tab 60 \tab 917 \tab 279 \tab - \tab 2904 \cr
   1998 \tab - \tab 1340 \tab 311 \tab 15 \tab 899 \tab 267 \tab - \tab 2832 \cr
   2000 \tab - \tab 1305 \tab 293 \tab 7 \tab 996 \tab 216 \tab - \tab 2817 \cr
   2002 \tab - \tab 1361 \tab 158 \tab 9 \tab 960 \tab 277 \tab - \tab 2765 \cr
   2004 \tab - \tab 1372 \tab 172 \tab 3 \tab 989 \tab 276 \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 1318 \tab 213 \tab 2 \tab 1206 \tab 253 \tab - \tab 4510 \cr
   2008 \tab - \tab 783 \tab 143 \tab - \tab 943 \tab 154 \tab - \tab 2023 \cr
   2010 \tab - \tab 889 \tab 102 \tab 1 \tab 855 \tab 197 \tab - \tab 2044 \cr
   2012 \tab - \tab 888 \tab 150 \tab 4 \tab 730 \tab 202 \tab - \tab 1974 \cr
   2014 \tab - \tab 1210 \tab 162 \tab 2 \tab 943 \tab 221 \tab - \tab 2538 \cr
   2016 \tab - \tab 1435 \tab 166 \tab 2 \tab 1016 \tab 248 \tab - \tab 2867 \cr
   2018 \tab - \tab 1121 \tab 151 \tab 2 \tab 887 \tab 187 \tab - \tab 2348 \cr
   2021 \tab - \tab 2039 \tab 14 \tab - \tab 1628 \tab 329 \tab 22 \tab 4032 \cr
   2022 \tab - \tab 1531 \tab 132 \tab 3 \tab 1586 \tab 286 \tab 6 \tab 3544 \cr
   Total \tab 17581 \tab 25600 \tab 4180 \tab 300 \tab 19696 \tab 5005 \tab 28 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
