% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_25.R
\name{natpark}
\alias{natpark}
\title{Parks and recreation}
\source{
General Social Survey https://gss.norc.org
}
\description{
natpark
}
\details{
Question 68. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. o. Parks and recreation.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 484 \tab 569 \tab 34 \tab 15 \tab 313 \tab 58 \tab - \tab 1473 \cr
   1985 \tab - \tab 889 \tab 65 \tab 4 \tab 468 \tab 108 \tab - \tab 1534 \cr
   1986 \tab - \tab 874 \tab 67 \tab 11 \tab 429 \tab 89 \tab - \tab 1470 \cr
   1987 \tab - \tab 1021 \tab 76 \tab 16 \tab 593 \tab 113 \tab - \tab 1819 \cr
   1988 \tab - \tab 894 \tab 72 \tab 13 \tab 421 \tab 81 \tab - \tab 1481 \cr
   1989 \tab - \tab 859 \tab 94 \tab 9 \tab 497 \tab 78 \tab - \tab 1537 \cr
   1990 \tab - \tab 783 \tab 82 \tab 19 \tab 412 \tab 76 \tab - \tab 1372 \cr
   1991 \tab - \tab 901 \tab 70 \tab 21 \tab 450 \tab 75 \tab - \tab 1517 \cr
   1993 \tab - \tab 920 \tab 90 \tab 15 \tab 480 \tab 101 \tab - \tab 1606 \cr
   1994 \tab - \tab 1769 \tab 116 \tab 45 \tab 874 \tab 188 \tab - \tab 2992 \cr
   1996 \tab - \tab 1616 \tab 139 \tab 61 \tab 920 \tab 168 \tab - \tab 2904 \cr
   1998 \tab - \tab 1564 \tab 124 \tab 8 \tab 963 \tab 173 \tab - \tab 2832 \cr
   2000 \tab - \tab 1547 \tab 116 \tab 8 \tab 991 \tab 155 \tab - \tab 2817 \cr
   2002 \tab - \tab 1598 \tab 76 \tab 4 \tab 940 \tab 147 \tab - \tab 2765 \cr
   2004 \tab - \tab 1685 \tab 54 \tab 3 \tab 879 \tab 191 \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 1734 \tab 70 \tab 2 \tab 996 \tab 190 \tab - \tab 4510 \cr
   2008 \tab - \tab 1252 \tab 45 \tab 1 \tab 617 \tab 108 \tab - \tab 2023 \cr
   2010 \tab - \tab 1201 \tab 41 \tab 1 \tab 672 \tab 129 \tab - \tab 2044 \cr
   2012 \tab - \tab 1185 \tab 51 \tab 1 \tab 620 \tab 117 \tab - \tab 1974 \cr
   2014 \tab - \tab 1540 \tab 47 \tab 2 \tab 790 \tab 159 \tab - \tab 2538 \cr
   2016 \tab - \tab 1692 \tab 48 \tab 2 \tab 952 \tab 173 \tab - \tab 2867 \cr
   2018 \tab - \tab 1346 \tab 60 \tab - \tab 839 \tab 103 \tab - \tab 2348 \cr
   2021 \tab - \tab 2412 \tab 10 \tab - \tab 1428 \tab 162 \tab 20 \tab 4032 \cr
   2022 \tab - \tab 1953 \tab 58 \tab - \tab 1347 \tab 178 \tab 8 \tab 3544 \cr
   Total \tab 17581 \tab 31804 \tab 1705 \tab 261 \tab 17891 \tab 3120 \tab 28 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
