% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_24.R
\name{natrace}
\alias{natrace}
\title{Improving the conditions of blacks}
\source{
General Social Survey https://gss.norc.org
}
\description{
natrace
}
\details{
Question 68. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. h. Improving the conditions of Blacks.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 590 \tab 97 \tab 5 \tab 488 \tab 324 \tab - \tab 1504 \cr
   1974 \tab - \tab 616 \tab 98 \tab 7 \tab 457 \tab 306 \tab - \tab 1484 \cr
   1975 \tab - \tab 616 \tab 114 \tab 4 \tab 400 \tab 356 \tab - \tab 1490 \cr
   1976 \tab - \tab 604 \tab 99 \tab 8 \tab 409 \tab 379 \tab - \tab 1499 \cr
   1977 \tab - \tab 644 \tab 123 \tab 5 \tab 383 \tab 375 \tab - \tab 1530 \cr
   1978 \tab - \tab 661 \tab 112 \tab 3 \tab 371 \tab 385 \tab - \tab 1532 \cr
   1980 \tab - \tab 645 \tab 115 \tab 6 \tab 353 \tab 349 \tab - \tab 1468 \cr
   1982 \tab - \tab 686 \tab 128 \tab 10 \tab 728 \tab 308 \tab - \tab 1860 \cr
   1983 \tab - \tab 692 \tab 145 \tab 7 \tab 466 \tab 289 \tab - \tab 1599 \cr
   1984 \tab 983 \tab 208 \tab 30 \tab 5 \tab 170 \tab 77 \tab - \tab 1473 \cr
   1985 \tab 783 \tab 319 \tab 50 \tab 2 \tab 230 \tab 150 \tab - \tab 1534 \cr
   1986 \tab 740 \tab 308 \tab 58 \tab 6 \tab 243 \tab 115 \tab - \tab 1470 \cr
   1987 \tab 1216 \tab 223 \tab 40 \tab 7 \tab 262 \tab 71 \tab - \tab 1819 \cr
   1988 \tab 763 \tab 293 \tab 59 \tab 4 \tab 248 \tab 114 \tab - \tab 1481 \cr
   1989 \tab 769 \tab 317 \tab 79 \tab 3 \tab 247 \tab 122 \tab - \tab 1537 \cr
   1990 \tab 698 \tab 271 \tab 52 \tab 4 \tab 247 \tab 100 \tab - \tab 1372 \cr
   1991 \tab 761 \tab 321 \tab 64 \tab 7 \tab 256 \tab 108 \tab - \tab 1517 \cr
   1993 \tab 807 \tab 315 \tab 73 \tab 6 \tab 284 \tab 121 \tab - \tab 1606 \cr
   1994 \tab 1484 \tab 605 \tab 130 \tab 6 \tab 462 \tab 305 \tab - \tab 2992 \cr
   1996 \tab 1463 \tab 558 \tab 145 \tab 4 \tab 457 \tab 277 \tab - \tab 2904 \cr
   1998 \tab 1451 \tab 555 \tab 126 \tab 6 \tab 469 \tab 225 \tab - \tab 2832 \cr
   2000 \tab 1408 \tab 563 \tab 163 \tab 10 \tab 469 \tab 204 \tab - \tab 2817 \cr
   2002 \tab 1407 \tab 617 \tab 87 \tab 12 \tab 412 \tab 230 \tab - \tab 2765 \cr
   2004 \tab 1411 \tab 636 \tab 115 \tab 10 \tab 441 \tab 199 \tab - \tab 2812 \cr
   2006 \tab 3026 \tab 602 \tab 124 \tab 12 \tab 521 \tab 225 \tab - \tab 4510 \cr
   2008 \tab 1019 \tab 420 \tab 113 \tab 10 \tab 350 \tab 111 \tab - \tab 2023 \cr
   2010 \tab 1030 \tab 439 \tab 108 \tab 7 \tab 306 \tab 154 \tab - \tab 2044 \cr
   2012 \tab 976 \tab 420 \tab 111 \tab 10 \tab 329 \tab 128 \tab - \tab 1974 \cr
   2014 \tab 1264 \tab 550 \tab 135 \tab 9 \tab 406 \tab 174 \tab - \tab 2538 \cr
   2016 \tab 1430 \tab 484 \tab 120 \tab 23 \tab 672 \tab 138 \tab - \tab 2867 \cr
   2018 \tab 1160 \tab 386 \tab 100 \tab 10 \tab 605 \tab 87 \tab - \tab 2348 \cr
   2021 \tab 2045 \tab 596 \tab 12 \tab 2 \tab 1020 \tab 343 \tab 14 \tab 4032 \cr
   2022 \tab 1764 \tab 515 \tab 76 \tab 10 \tab 941 \tab 229 \tab 9 \tab 3544 \cr
   Total \tab 31471 \tab 16275 \tab 3201 \tab 240 \tab 14102 \tab 7078 \tab 23 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
