% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_26.R
\name{natracey}
\alias{natracey}
\title{natracey}
\source{
General Social Survey https://gss.norc.org
}
\description{
Assistance to blacks:version y
}
\details{
Question 69. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. h. Assistance to blacks.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 974 \tab 203 \tab 48 \tab 13 \tab 119 \tab 116 \tab - \tab 1473 \cr
   1985 \tab 751 \tab 314 \tab 74 \tab 2 \tab 200 \tab 193 \tab - \tab 1534 \cr
   1986 \tab 730 \tab 331 \tab 45 \tab 7 \tab 166 \tab 191 \tab - \tab 1470 \cr
   1987 \tab 603 \tab 451 \tab 101 \tab 11 \tab 433 \tab 220 \tab - \tab 1819 \cr
   1988 \tab 718 \tab 301 \tab 90 \tab 10 \tab 191 \tab 171 \tab - \tab 1481 \cr
   1989 \tab 768 \tab 308 \tab 79 \tab 8 \tab 207 \tab 167 \tab - \tab 1537 \cr
   1990 \tab 674 \tab 291 \tab 53 \tab 15 \tab 197 \tab 142 \tab - \tab 1372 \cr
   1991 \tab 756 \tab 299 \tab 76 \tab 18 \tab 227 \tab 141 \tab - \tab 1517 \cr
   1993 \tab 799 \tab 306 \tab 108 \tab 14 \tab 188 \tab 191 \tab - \tab 1606 \cr
   1994 \tab 1508 \tab 565 \tab 156 \tab 36 \tab 349 \tab 378 \tab - \tab 2992 \cr
   1996 \tab 1441 \tab 544 \tab 187 \tab 55 \tab 334 \tab 343 \tab - \tab 2904 \cr
   1998 \tab 1381 \tab 565 \tab 180 \tab 9 \tab 381 \tab 316 \tab - \tab 2832 \cr
   2000 \tab 1409 \tab 522 \tab 172 \tab 9 \tab 425 \tab 280 \tab - \tab 2817 \cr
   2002 \tab 1358 \tab 591 \tab 133 \tab 11 \tab 338 \tab 334 \tab - \tab 2765 \cr
   2004 \tab 1401 \tab 615 \tab 165 \tab 18 \tab 363 \tab 250 \tab - \tab 2812 \cr
   2006 \tab 3002 \tab 608 \tab 195 \tab 12 \tab 378 \tab 315 \tab - \tab 4510 \cr
   2008 \tab 1004 \tab 404 \tab 118 \tab 9 \tab 279 \tab 209 \tab - \tab 2023 \cr
   2010 \tab 1014 \tab 427 \tab 147 \tab 14 \tab 252 \tab 190 \tab - \tab 2044 \cr
   2012 \tab 998 \tab 407 \tab 156 \tab 12 \tab 206 \tab 195 \tab - \tab 1974 \cr
   2014 \tab 1274 \tab 510 \tab 161 \tab 15 \tab 316 \tab 262 \tab - \tab 2538 \cr
   2016 \tab 1437 \tab 535 \tab 139 \tab 6 \tab 534 \tab 216 \tab - \tab 2867 \cr
   2018 \tab 1188 \tab 379 \tab 128 \tab 6 \tab 507 \tab 140 \tab - \tab 2348 \cr
   2021 \tab 1987 \tab 692 \tab 16 \tab 3 \tab 930 \tab 384 \tab 20 \tab 4032 \cr
   2022 \tab 1780 \tab 566 \tab 97 \tab 8 \tab 808 \tab 278 \tab 7 \tab 3544 \cr
   Total \tab 44534 \tab 10734 \tab 2824 \tab 321 \tab 8328 \tab 5622 \tab 27 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
