% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_25.R
\name{natsoc}
\alias{natsoc}
\title{natsoc}
\source{
General Social Survey https://gss.norc.org
}
\description{
Social security
}
\details{
Question 68. We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. m. Social Security.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 484 \tab 344 \tab 46 \tab 21 \tab 493 \tab 85 \tab - \tab 1473 \cr
   1985 \tab - \tab 586 \tab 46 \tab 7 \tab 794 \tab 101 \tab - \tab 1534 \cr
   1986 \tab - \tab 519 \tab 37 \tab 14 \tab 810 \tab 90 \tab - \tab 1470 \cr
   1987 \tab - \tab 570 \tab 74 \tab 19 \tab 1046 \tab 110 \tab - \tab 1819 \cr
   1988 \tab - \tab 553 \tab 61 \tab 14 \tab 772 \tab 81 \tab - \tab 1481 \cr
   1989 \tab - \tab 562 \tab 73 \tab 10 \tab 826 \tab 66 \tab - \tab 1537 \cr
   1990 \tab - \tab 538 \tab 70 \tab 21 \tab 667 \tab 76 \tab - \tab 1372 \cr
   1991 \tab - \tab 580 \tab 81 \tab 22 \tab 774 \tab 60 \tab - \tab 1517 \cr
   1993 \tab - \tab 686 \tab 99 \tab 15 \tab 693 \tab 113 \tab - \tab 1606 \cr
   1994 \tab - \tab 1221 \tab 151 \tab 44 \tab 1371 \tab 205 \tab - \tab 2992 \cr
   1996 \tab - \tab 1066 \tab 163 \tab 63 \tab 1388 \tab 224 \tab - \tab 2904 \cr
   1998 \tab - \tab 903 \tab 161 \tab 13 \tab 1578 \tab 177 \tab - \tab 2832 \cr
   2000 \tab - \tab 897 \tab 157 \tab 6 \tab 1626 \tab 131 \tab - \tab 2817 \cr
   2002 \tab - \tab 920 \tab 99 \tab 5 \tab 1618 \tab 123 \tab - \tab 2765 \cr
   2004 \tab - \tab 778 \tab 103 \tab 7 \tab 1772 \tab 152 \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 891 \tab 124 \tab 4 \tab 1835 \tab 138 \tab - \tab 4510 \cr
   2008 \tab - \tab 608 \tab 93 \tab 4 \tab 1206 \tab 112 \tab - \tab 2023 \cr
   2010 \tab - \tab 677 \tab 91 \tab 5 \tab 1126 \tab 145 \tab - \tab 2044 \cr
   2012 \tab - \tab 683 \tab 94 \tab 3 \tab 1048 \tab 146 \tab - \tab 1974 \cr
   2014 \tab - \tab 906 \tab 124 \tab 3 \tab 1353 \tab 152 \tab - \tab 2538 \cr
   2016 \tab - \tab 922 \tab 108 \tab 3 \tab 1690 \tab 144 \tab - \tab 2867 \cr
   2018 \tab - \tab 748 \tab 115 \tab 3 \tab 1381 \tab 101 \tab - \tab 2348 \cr
   2021 \tab - \tab 1518 \tab 10 \tab - \tab 2286 \tab 198 \tab 20 \tab 4032 \cr
   2022 \tab - \tab 1051 \tab 107 \tab 3 \tab 2213 \tab 161 \tab 9 \tab 3544 \cr
   Total \tab 17581 \tab 18727 \tab 2287 \tab 309 \tab 30366 \tab 3091 \tab 29 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
