% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_77.R
\name{news}
\alias{news}
\title{How often does r read newspaper}
\source{
General Social Survey https://gss.norc.org
}
\description{
news
}
\details{
Question 241. How often do you read the newspaper - every day, a few times a week, once a week, less than once a week, or never?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab a few times a week \tab every day \tab less than once a week \tab never \tab no answer \tab once a week \tab iap \tab skipped on web \tab don't know \tab Total \cr
   1972 \tab 241 \tab 1105 \tab 69 \tab 68 \tab 2 \tab 128 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \tab - \tab - \tab 1484 \cr
   1975 \tab 235 \tab 981 \tab 81 \tab 65 \tab 2 \tab 126 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \tab - \tab - \tab 1499 \cr
   1977 \tab 254 \tab 952 \tab 100 \tab 73 \tab 3 \tab 148 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 308 \tab 874 \tab 113 \tab 79 \tab 4 \tab 154 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \tab - \tab - \tab 1468 \cr
   1982 \tab 425 \tab 961 \tab 138 \tab 112 \tab 6 \tab 218 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 331 \tab 891 \tab 127 \tab 80 \tab - \tab 170 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \tab - \tab - \tab 1473 \cr
   1985 \tab 324 \tab 807 \tab 116 \tab 87 \tab 4 \tab 196 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 289 \tab 789 \tab 116 \tab 90 \tab 2 \tab 184 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 391 \tab 949 \tab 130 \tab 110 \tab 8 \tab 231 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 235 \tab 500 \tab 87 \tab 48 \tab - \tab 118 \tab 493 \tab - \tab - \tab 1481 \cr
   1989 \tab 248 \tab 502 \tab 76 \tab 49 \tab 1 \tab 130 \tab 531 \tab - \tab - \tab 1537 \cr
   1990 \tab 205 \tab 488 \tab 89 \tab 45 \tab 1 \tab 100 \tab 444 \tab - \tab - \tab 1372 \cr
   1991 \tab 220 \tab 525 \tab 87 \tab 51 \tab 2 \tab 139 \tab 493 \tab - \tab - \tab 1517 \cr
   1993 \tab 282 \tab 497 \tab 96 \tab 64 \tab 5 \tab 136 \tab 526 \tab - \tab - \tab 1606 \cr
   1994 \tab 452 \tab 969 \tab 196 \tab 76 \tab - \tab 261 \tab 1038 \tab - \tab - \tab 2992 \cr
   1996 \tab 476 \tab 825 \tab 225 \tab 120 \tab 3 \tab 311 \tab 944 \tab - \tab - \tab 2904 \cr
   1998 \tab 420 \tab 805 \tab 202 \tab 149 \tab 1 \tab 294 \tab 961 \tab - \tab - \tab 2832 \cr
   2000 \tab 459 \tab 698 \tab 277 \tab 162 \tab - \tab 277 \tab 944 \tab - \tab - \tab 2817 \cr
   2002 \tab 219 \tab 369 \tab 121 \tab 90 \tab - \tab 109 \tab 1857 \tab - \tab - \tab 2765 \cr
   2004 \tab 222 \tab 358 \tab 121 \tab 71 \tab - \tab 134 \tab 1906 \tab - \tab - \tab 2812 \cr
   2006 \tab 611 \tab 945 \tab 404 \tab 350 \tab - \tab 418 \tab 1782 \tab - \tab - \tab 4510 \cr
   2008 \tab 300 \tab 431 \tab 200 \tab 191 \tab - \tab 207 \tab 694 \tab - \tab - \tab 2023 \cr
   2010 \tab 277 \tab 449 \tab 220 \tab 269 \tab - \tab 215 \tab 614 \tab - \tab - \tab 2044 \cr
   2012 \tab 230 \tab 353 \tab 217 \tab 306 \tab 1 \tab 195 \tab 672 \tab - \tab - \tab 1974 \cr
   2014 \tab 260 \tab 417 \tab 271 \tab 481 \tab - \tab 246 \tab 863 \tab - \tab - \tab 2538 \cr
   2016 \tab 270 \tab 398 \tab 300 \tab 688 \tab 2 \tab 229 \tab 980 \tab - \tab - \tab 2867 \cr
   2018 \tab 174 \tab 312 \tab 211 \tab 678 \tab - \tab 184 \tab 789 \tab - \tab - \tab 2348 \cr
   2021 \tab 361 \tab 642 \tab 452 \tab 1001 \tab - \tab 240 \tab 1329 \tab 7 \tab - \tab 4032 \cr
   2022 \tab 267 \tab 383 \tab 329 \tab 1191 \tab 1 \tab 195 \tab 1168 \tab 6 \tab 4 \tab 3544 \cr
   Total \tab 8986 \tab 19175 \tab 5171 \tab 6844 \tab 48 \tab 5693 \tab 26456 \tab 13 \tab 4 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} every day
\item \code{2} a few times a week
\item \code{3} once a week
\item \code{4} less than once a week
\item \code{5} never
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
