% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_366.R
\name{newsfrom}
\alias{newsfrom}
\title{Main source of information about events in the news}
\source{
General Social Survey https://gss.norc.org
}
\description{
newsfrom
}
\details{
Question 1032. We are interested in how people get information about events in the news. Where do you get most of your information about current news events ~ newspapers, magazines, the Internet, books or other printed materials, TV, radio, government agencies, family, friends, colleagues, or some other source?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab books/other printed material \tab don't know \tab family \tab friends/colleagues \tab government agencies \tab magazines \tab newspaper \tab no answer \tab other \tab radio \tab the internet \tab tv \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2646 \tab 9 \tab 3 \tab 33 \tab 33 \tab 3 \tab 20 \tab 446 \tab 2 \tab 15 \tab 131 \tab 253 \tab 916 \tab 4510 \cr
   2008 \tab 518 \tab 8 \tab 2 \tab 21 \tab 20 \tab 2 \tab 16 \tab 286 \tab 3 \tab 3 \tab 93 \tab 310 \tab 741 \tab 2023 \cr
   2010 \tab 1761 \tab 3 \tab - \tab 1 \tab 7 \tab 1 \tab 1 \tab 45 \tab - \tab - \tab 19 \tab 62 \tab 144 \tab 2044 \cr
   2012 \tab 972 \tab 5 \tab 4 \tab 17 \tab 13 \tab 1 \tab 6 \tab 121 \tab 4 \tab 2 \tab 54 \tab 299 \tab 476 \tab 1974 \cr
   2014 \tab 1299 \tab 4 \tab 1 \tab 25 \tab 20 \tab 3 \tab 6 \tab 134 \tab - \tab 1 \tab 75 \tab 420 \tab 550 \tab 2538 \cr
   2016 \tab 1477 \tab 4 \tab - \tab 29 \tab 25 \tab 1 \tab 7 \tab 110 \tab 4 \tab 12 \tab 65 \tab 601 \tab 532 \tab 2867 \cr
   2018 \tab 1173 \tab 5 \tab 1 \tab 30 \tab 22 \tab 4 \tab 2 \tab 79 \tab - \tab 2 \tab 69 \tab 539 \tab 422 \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 63932 \tab 38 \tab 11 \tab 156 \tab 140 \tab 15 \tab 58 \tab 1221 \tab 13 \tab 35 \tab 506 \tab 2484 \tab 3781 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} newspaper
\item \code{2} magazines
\item \code{3} the internet
\item \code{4} books/other printed material
\item \code{5} tv
\item \code{6} radio
\item \code{7} government agencies
\item \code{8} family
\item \code{9} friends/colleagues
\item \code{10} other
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
