% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_57.R
\name{obey}
\alias{obey}
\title{obey}
\source{
General Social Survey https://gss.norc.org
}
\description{
To obey
}
\details{
Question 170. If you had to choose, which thing on this list would you pick as the most important for a child to learn to prepare him or her for life? a. Which comes next in importantc? b. Which comes third? c. Which comes fourth? TO OBEY
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab fifth \tab first \tab fourth \tab no answer \tab second \tab third \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 730 \tab 96 \tab 168 \tab 235 \tab 8 \tab 116 \tab 117 \tab 1470 \cr
   1987 \tab - \tab 326 \tab 433 \tab 525 \tab 19 \tab 250 \tab 266 \tab 1819 \cr
   1988 \tab 484 \tab 152 \tab 226 \tab 286 \tab 20 \tab 122 \tab 191 \tab 1481 \cr
   1989 \tab 502 \tab 167 \tab 199 \tab 329 \tab 35 \tab 137 \tab 168 \tab 1537 \cr
   1990 \tab 473 \tab 147 \tab 165 \tab 285 \tab 28 \tab 108 \tab 166 \tab 1372 \cr
   1991 \tab 500 \tab 179 \tab 195 \tab 343 \tab 35 \tab 123 \tab 142 \tab 1517 \cr
   1993 \tab 549 \tab 159 \tab 191 \tab 361 \tab 26 \tab 146 \tab 174 \tab 1606 \cr
   1994 \tab 981 \tab 273 \tab 357 \tab 759 \tab 33 \tab 261 \tab 328 \tab 2992 \cr
   1996 \tab 979 \tab 279 \tab 341 \tab 689 \tab 30 \tab 250 \tab 336 \tab 2904 \cr
   1998 \tab 921 \tab 288 \tab 355 \tab 680 \tab 31 \tab 237 \tab 320 \tab 2832 \cr
   2000 \tab 921 \tab 283 \tab 389 \tab 662 \tab 33 \tab 230 \tab 299 \tab 2817 \cr
   2002 \tab 1853 \tab 144 \tab 131 \tab 367 \tab 13 \tab 112 \tab 145 \tab 2765 \cr
   2004 \tab 1936 \tab 127 \tab 107 \tab 365 \tab 8 \tab 123 \tab 146 \tab 2812 \cr
   2006 \tab 3508 \tab 133 \tab 184 \tab 389 \tab 7 \tab 119 \tab 170 \tab 4510 \cr
   2008 \tab 658 \tab 206 \tab 212 \tab 555 \tab 14 \tab 155 \tab 223 \tab 2023 \cr
   2010 \tab 667 \tab 204 \tab 193 \tab 537 \tab 28 \tab 170 \tab 245 \tab 2044 \cr
   2012 \tab 636 \tab 210 \tab 182 \tab 537 \tab 15 \tab 156 \tab 238 \tab 1974 \cr
   2014 \tab 851 \tab 292 \tab 204 \tab 718 \tab 17 \tab 203 \tab 253 \tab 2538 \cr
   2016 \tab 911 \tab 327 \tab 239 \tab 847 \tab 22 \tab 211 \tab 310 \tab 2867 \cr
   2018 \tab 785 \tab 260 \tab 167 \tab 706 \tab 21 \tab 169 \tab 240 \tab 2348 \cr
   2021 \tab 1360 \tab 674 \tab 127 \tab 1318 \tab 99 \tab 160 \tab 294 \tab 4032 \cr
   2022 \tab 1173 \tab 594 \tab 169 \tab 1085 \tab 77 \tab 176 \tab 270 \tab 3544 \cr
   Total \tab 39964 \tab 5520 \tab 4934 \tab 12578 \tab 619 \tab 3734 \tab 5041 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} first
\item \code{2} second
\item \code{3} third
\item \code{4} fourth
\item \code{5} fifth
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
