% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_02.R
\name{occindv}
\alias{occindv}
\title{occindv}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's occ/ind verbatim retrieved
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab no verbatims for r \tab verbatims for r \tab Total \cr
   1972 \tab 911 \tab 71 \tab 631 \tab 1613 \cr
   1973 \tab 270 \tab 114 \tab 1120 \tab 1504 \cr
   1974 \tab 267 \tab 110 \tab 1107 \tab 1484 \cr
   1975 \tab 514 \tab 91 \tab 885 \tab 1490 \cr
   1976 \tab 269 \tab 111 \tab 1119 \tab 1499 \cr
   1977 \tab 713 \tab 55 \tab 762 \tab 1530 \cr
   1978 \tab 14 \tab 113 \tab 1405 \tab 1532 \cr
   1980 \tab 175 \tab 94 \tab 1199 \tab 1468 \cr
   1982 \tab 164 \tab 96 \tab 1600 \tab 1860 \cr
   1983 \tab 124 \tab 87 \tab 1388 \tab 1599 \cr
   1984 \tab 15 \tab 81 \tab 1377 \tab 1473 \cr
   1985 \tab 238 \tab 68 \tab 1228 \tab 1534 \cr
   1986 \tab 45 \tab 82 \tab 1343 \tab 1470 \cr
   1987 \tab 308 \tab 61 \tab 1450 \tab 1819 \cr
   1988 \tab 21 \tab 75 \tab 1385 \tab 1481 \cr
   1989 \tab 117 \tab 70 \tab 1350 \tab 1537 \cr
   1990 \tab 14 \tab 61 \tab 1297 \tab 1372 \cr
   1991 \tab 18 \tab 83 \tab 1416 \tab 1517 \cr
   1993 \tab 61 \tab 59 \tab 1486 \tab 1606 \cr
   1994 \tab 54 \tab 118 \tab 2820 \tab 2992 \cr
   1996 \tab 21 \tab 104 \tab 2779 \tab 2904 \cr
   1998 \tab 31 \tab 123 \tab 2678 \tab 2832 \cr
   2000 \tab - \tab 135 \tab 2682 \tab 2817 \cr
   2002 \tab - \tab 100 \tab 2665 \tab 2765 \cr
   2004 \tab - \tab 144 \tab 2668 \tab 2812 \cr
   2006 \tab - \tab 209 \tab 4301 \tab 4510 \cr
   2008 \tab - \tab 66 \tab 1957 \tab 2023 \cr
   2010 \tab - \tab 103 \tab 1941 \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab 3544 \cr
   Total \tab 21667 \tab 2684 \tab 48039 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} no verbatims for r
\item \code{1} verbatims for r
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
