% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_581.R
\name{old1}
\alias{old1}
\title{old1}
\source{
General Social Survey https://gss.norc.org
}
\description{
Age of 1st person
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. c. How old was \link{PERSON} on his/her last birthday?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 76 \tab 77 \tab 78 \tab 79 \tab 80 \tab 81 \tab 82 \tab 83 \tab 84 \tab 85 \tab 87 \tab 88 \tab 89 \tab don't know \tab no answer \tab 86 \tab 90 \tab 92 \tab 93 \tab 91 \tab 97 or older \tab 17 \tab 94 \tab 95 \tab 96 \tab not available in this release \tab 0 \tab 10 \tab 15 \tab 150 \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 7 \tab 8 \tab 16 \tab 16 \tab 15 \tab 28 \tab 39 \tab 29 \tab 31 \tab 38 \tab 33 \tab 29 \tab 27 \tab 34 \tab 33 \tab 26 \tab 26 \tab 16 \tab 23 \tab 36 \tab 23 \tab 26 \tab 31 \tab 26 \tab 28 \tab 26 \tab 30 \tab 32 \tab 18 \tab 24 \tab 26 \tab 31 \tab 30 \tab 28 \tab 33 \tab 31 \tab 23 \tab 26 \tab 27 \tab 24 \tab 28 \tab 15 \tab 25 \tab 17 \tab 31 \tab 20 \tab 23 \tab 29 \tab 22 \tab 23 \tab 22 \tab 28 \tab 15 \tab 8 \tab 13 \tab 12 \tab 9 \tab 9 \tab 13 \tab 11 \tab 8 \tab 9 \tab 12 \tab 8 \tab 5 \tab 4 \tab 3 \tab 3 \tab 2 \tab 1 \tab 2 \tab 3 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 755 \tab 4 \tab 5 \tab 4 \tab 6 \tab 16 \tab 16 \tab 12 \tab 14 \tab 20 \tab 21 \tab 16 \tab 19 \tab 15 \tab 21 \tab 16 \tab 17 \tab 15 \tab 15 \tab 12 \tab 15 \tab 18 \tab 12 \tab 15 \tab 12 \tab 9 \tab 12 \tab 8 \tab 10 \tab 14 \tab 7 \tab 9 \tab 9 \tab 11 \tab 15 \tab 15 \tab 11 \tab 8 \tab 9 \tab 13 \tab 9 \tab 13 \tab 10 \tab 16 \tab 17 \tab 7 \tab 11 \tab 9 \tab 6 \tab 15 \tab 13 \tab 7 \tab 15 \tab 5 \tab 3 \tab 8 \tab 9 \tab 6 \tab 10 \tab 5 \tab 7 \tab 8 \tab 5 \tab 2 \tab 4 \tab 4 \tab 2 \tab 3 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 4 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 2 \tab 6 \tab 9 \tab 21 \tab 15 \tab 26 \tab 31 \tab 32 \tab 32 \tab 31 \tab 33 \tab 38 \tab 30 \tab 17 \tab 30 \tab 34 \tab 29 \tab 33 \tab 29 \tab 25 \tab 25 \tab 24 \tab 24 \tab 32 \tab 33 \tab 20 \tab 31 \tab 27 \tab 29 \tab 32 \tab 32 \tab 40 \tab 34 \tab 23 \tab 35 \tab 27 \tab 32 \tab 33 \tab 33 \tab 29 \tab 34 \tab 34 \tab 19 \tab 24 \tab 32 \tab 18 \tab 21 \tab 20 \tab 18 \tab 29 \tab 15 \tab 15 \tab 12 \tab 8 \tab 23 \tab 18 \tab 11 \tab 14 \tab 14 \tab 9 \tab 5 \tab 4 \tab 6 \tab 3 \tab 5 \tab 3 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 3 \tab 4 \tab 13 \tab 18 \tab 22 \tab 35 \tab 28 \tab 41 \tab 33 \tab 39 \tab 33 \tab 37 \tab 47 \tab 48 \tab 29 \tab 37 \tab 38 \tab 28 \tab 33 \tab 24 \tab 29 \tab 31 \tab 34 \tab 23 \tab 32 \tab 32 \tab 16 \tab 19 \tab 19 \tab 21 \tab 33 \tab 20 \tab 23 \tab 28 \tab 30 \tab 26 \tab 18 \tab 31 \tab 28 \tab 32 \tab 22 \tab 22 \tab 20 \tab 24 \tab 19 \tab 19 \tab 19 \tab 25 \tab 19 \tab 17 \tab 11 \tab 20 \tab 20 \tab 8 \tab 18 \tab 18 \tab 11 \tab 15 \tab 9 \tab 12 \tab 14 \tab 6 \tab 9 \tab 3 \tab 5 \tab 6 \tab 4 \tab 6 \tab 2 \tab - \tab 3 \tab - \tab 6 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 6 \tab 8 \tab 14 \tab 20 \tab 21 \tab 37 \tab 22 \tab 27 \tab 45 \tab 19 \tab 34 \tab 38 \tab 33 \tab 28 \tab 51 \tab 29 \tab 29 \tab 38 \tab 31 \tab 32 \tab 12 \tab 29 \tab 26 \tab 21 \tab 21 \tab 30 \tab 16 \tab 18 \tab 21 \tab 18 \tab 25 \tab 23 \tab 28 \tab 22 \tab 27 \tab 26 \tab 22 \tab 26 \tab 28 \tab 24 \tab 32 \tab 28 \tab 17 \tab 15 \tab 24 \tab 16 \tab 19 \tab 29 \tab 18 \tab 24 \tab 15 \tab 22 \tab 22 \tab 15 \tab 18 \tab 14 \tab 10 \tab 11 \tab 14 \tab 9 \tab 8 \tab 9 \tab 8 \tab 4 \tab 6 \tab 5 \tab 4 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 3 \tab 6 \tab 2 \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 1 \tab 6 \tab 12 \tab 24 \tab 22 \tab 24 \tab 40 \tab 39 \tab 50 \tab 51 \tab 47 \tab 50 \tab 45 \tab 38 \tab 39 \tab 18 \tab 41 \tab 39 \tab 40 \tab 35 \tab 40 \tab 30 \tab 36 \tab 29 \tab 26 \tab 31 \tab 27 \tab 27 \tab 26 \tab 20 \tab 27 \tab 33 \tab 27 \tab 31 \tab 29 \tab 32 \tab 28 \tab 43 \tab 33 \tab 37 \tab 34 \tab 29 \tab 41 \tab 24 \tab 34 \tab 18 \tab 28 \tab 24 \tab 27 \tab 38 \tab 18 \tab 28 \tab 24 \tab 25 \tab 20 \tab 14 \tab 17 \tab 13 \tab 8 \tab 15 \tab 12 \tab 10 \tab 7 \tab 8 \tab 12 \tab 5 \tab 11 \tab 5 \tab 5 \tab 2 \tab 1 \tab - \tab 18 \tab 5 \tab - \tab 2 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 2 \tab 5 \tab 7 \tab 12 \tab 20 \tab 33 \tab 33 \tab 37 \tab 48 \tab 50 \tab 34 \tab 33 \tab 47 \tab 38 \tab 39 \tab 34 \tab 35 \tab 37 \tab 38 \tab 34 \tab 29 \tab 31 \tab 30 \tab 27 \tab 34 \tab 32 \tab 24 \tab 25 \tab 21 \tab 24 \tab 20 \tab 27 \tab 24 \tab 19 \tab 27 \tab 23 \tab 24 \tab 27 \tab 28 \tab 24 \tab 24 \tab 29 \tab 27 \tab 15 \tab 24 \tab 17 \tab 26 \tab 18 \tab 21 \tab 25 \tab 22 \tab 18 \tab 17 \tab 18 \tab 22 \tab 13 \tab 19 \tab 11 \tab 10 \tab 11 \tab 8 \tab 8 \tab 10 \tab 6 \tab 11 \tab 8 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 9 \tab 3 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab - \tab 9 \tab 12 \tab 17 \tab 28 \tab 31 \tab 31 \tab 32 \tab 46 \tab 28 \tab 33 \tab 31 \tab 26 \tab 31 \tab 41 \tab 26 \tab 38 \tab 41 \tab 29 \tab 48 \tab 36 \tab 31 \tab 27 \tab 26 \tab 23 \tab 32 \tab 18 \tab 20 \tab 25 \tab 24 \tab 17 \tab 20 \tab 27 \tab 18 \tab 20 \tab 26 \tab 21 \tab 24 \tab 19 \tab 20 \tab 24 \tab 21 \tab 26 \tab 24 \tab 21 \tab 17 \tab 27 \tab 17 \tab 18 \tab 22 \tab 12 \tab 21 \tab 17 \tab 17 \tab 10 \tab 13 \tab 15 \tab 16 \tab 9 \tab 8 \tab 3 \tab 5 \tab 10 \tab 7 \tab 6 \tab 6 \tab 5 \tab 5 \tab 4 \tab 3 \tab - \tab - \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 3 \tab 10 \tab 8 \tab 16 \tab 28 \tab 19 \tab 33 \tab 36 \tab 33 \tab 31 \tab 35 \tab 27 \tab 32 \tab 35 \tab 38 \tab 33 \tab 33 \tab 34 \tab 29 \tab 45 \tab 28 \tab 27 \tab 37 \tab 23 \tab 29 \tab 30 \tab 22 \tab 29 \tab 18 \tab 18 \tab 22 \tab 26 \tab 37 \tab 22 \tab 18 \tab 23 \tab 20 \tab 20 \tab 19 \tab 35 \tab 24 \tab 24 \tab 23 \tab 23 \tab 29 \tab 22 \tab 19 \tab 26 \tab 18 \tab 24 \tab 23 \tab 17 \tab 21 \tab 12 \tab 14 \tab 15 \tab 13 \tab 12 \tab 16 \tab 15 \tab 10 \tab 8 \tab 7 \tab 11 \tab 7 \tab 3 \tab 11 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 7 \tab 7 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab 5 \tab 6 \tab 18 \tab 15 \tab 13 \tab 27 \tab 37 \tab 28 \tab 42 \tab 33 \tab 41 \tab 32 \tab 33 \tab 25 \tab 36 \tab 36 \tab 30 \tab 30 \tab 24 \tab 34 \tab 36 \tab 36 \tab 21 \tab 31 \tab 34 \tab 21 \tab 30 \tab 21 \tab 28 \tab 14 \tab 25 \tab 22 \tab 14 \tab 16 \tab 18 \tab 21 \tab 20 \tab 27 \tab 19 \tab 19 \tab 23 \tab 21 \tab 25 \tab 18 \tab 28 \tab 28 \tab 25 \tab 17 \tab 18 \tab 19 \tab 21 \tab 19 \tab 11 \tab 19 \tab 15 \tab 11 \tab 12 \tab 19 \tab 11 \tab 12 \tab 7 \tab 13 \tab 10 \tab 6 \tab 8 \tab 3 \tab 3 \tab 4 \tab 1 \tab 2 \tab - \tab 17 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 3 \tab 7 \tab 13 \tab 22 \tab 28 \tab 36 \tab 27 \tab 38 \tab 28 \tab 33 \tab 41 \tab 45 \tab 45 \tab 41 \tab 35 \tab 51 \tab 43 \tab 41 \tab 41 \tab 42 \tab 38 \tab 46 \tab 43 \tab 31 \tab 36 \tab 34 \tab 45 \tab 38 \tab 31 \tab 32 \tab 26 \tab 28 \tab 28 \tab 28 \tab 24 \tab 30 \tab 32 \tab 22 \tab 30 \tab 21 \tab 27 \tab 22 \tab 24 \tab 20 \tab 20 \tab 20 \tab 24 \tab 33 \tab 16 \tab 33 \tab 18 \tab 22 \tab 22 \tab 15 \tab 16 \tab 18 \tab 24 \tab 16 \tab 11 \tab 16 \tab 12 \tab 7 \tab 9 \tab 7 \tab 7 \tab 5 \tab 4 \tab 6 \tab 6 \tab 4 \tab 2 \tab 1 \tab 19 \tab 3 \tab 4 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 3 \tab 10 \tab 19 \tab 19 \tab 18 \tab 28 \tab 26 \tab 28 \tab 28 \tab 37 \tab 34 \tab 31 \tab 30 \tab 23 \tab 32 \tab 21 \tab 41 \tab 32 \tab 40 \tab 45 \tab 26 \tab 37 \tab 33 \tab 31 \tab 31 \tab 29 \tab 29 \tab 24 \tab 18 \tab 18 \tab 16 \tab 21 \tab 29 \tab 13 \tab 21 \tab 23 \tab 12 \tab 15 \tab 16 \tab 16 \tab 25 \tab 24 \tab 18 \tab 20 \tab 21 \tab 21 \tab 15 \tab 19 \tab 24 \tab 34 \tab 16 \tab 18 \tab 23 \tab 18 \tab 20 \tab 14 \tab 19 \tab 12 \tab 10 \tab 16 \tab 13 \tab 7 \tab 5 \tab 8 \tab 9 \tab 2 \tab 4 \tab 2 \tab 2 \tab 5 \tab 3 \tab 1 \tab 20 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 2 \tab 2 \tab 18 \tab 12 \tab 18 \tab 14 \tab 30 \tab 25 \tab 28 \tab 33 \tab 34 \tab 22 \tab 36 \tab 36 \tab 32 \tab 29 \tab 35 \tab 46 \tab 43 \tab 38 \tab 30 \tab 33 \tab 35 \tab 29 \tab 37 \tab 25 \tab 23 \tab 38 \tab 34 \tab 33 \tab 17 \tab 27 \tab 37 \tab 22 \tab 18 \tab 15 \tab 21 \tab 25 \tab 17 \tab 23 \tab 16 \tab 18 \tab 16 \tab 22 \tab 15 \tab 21 \tab 14 \tab 33 \tab 15 \tab 21 \tab 28 \tab 27 \tab 16 \tab 20 \tab 27 \tab 12 \tab 15 \tab 11 \tab 15 \tab 13 \tab 13 \tab 8 \tab 10 \tab 13 \tab 3 \tab 6 \tab 4 \tab 4 \tab 3 \tab 3 \tab 4 \tab - \tab 12 \tab 4 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 2 \tab 5 \tab 9 \tab 16 \tab 15 \tab 21 \tab 21 \tab 12 \tab 19 \tab 28 \tab 32 \tab 30 \tab 26 \tab 34 \tab 32 \tab 34 \tab 42 \tab 36 \tab 33 \tab 33 \tab 28 \tab 26 \tab 31 \tab 25 \tab 31 \tab 33 \tab 31 \tab 22 \tab 28 \tab 24 \tab 29 \tab 12 \tab 20 \tab 11 \tab 23 \tab 13 \tab 15 \tab 19 \tab 21 \tab 21 \tab 10 \tab 12 \tab 28 \tab 18 \tab 19 \tab 20 \tab 18 \tab 26 \tab 20 \tab 19 \tab 16 \tab 15 \tab 8 \tab 16 \tab 16 \tab 19 \tab 11 \tab 17 \tab 19 \tab 13 \tab 10 \tab 10 \tab 8 \tab 7 \tab 9 \tab 9 \tab 6 \tab 1 \tab 2 \tab 5 \tab 4 \tab 3 \tab 7 \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 3 \tab 8 \tab 12 \tab 24 \tab 21 \tab 19 \tab 28 \tab 23 \tab 25 \tab 22 \tab 32 \tab 25 \tab 29 \tab 40 \tab 44 \tab 33 \tab 42 \tab 33 \tab 28 \tab 35 \tab 41 \tab 39 \tab 43 \tab 32 \tab 42 \tab 32 \tab 24 \tab 23 \tab 36 \tab 23 \tab 24 \tab 19 \tab 15 \tab 22 \tab 29 \tab 14 \tab 23 \tab 19 \tab 17 \tab 13 \tab 13 \tab 17 \tab 29 \tab 20 \tab 22 \tab 12 \tab 20 \tab 15 \tab 14 \tab 21 \tab 22 \tab 20 \tab 19 \tab 17 \tab 27 \tab 14 \tab 24 \tab 10 \tab 9 \tab 10 \tab 11 \tab 15 \tab 10 \tab 13 \tab 6 \tab 6 \tab 5 \tab 6 \tab 4 \tab 1 \tab 2 \tab - \tab 16 \tab 4 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 4 \tab 5 \tab 6 \tab 12 \tab 16 \tab 17 \tab 23 \tab 22 \tab 24 \tab 36 \tab 31 \tab 27 \tab 34 \tab 33 \tab 29 \tab 44 \tab 34 \tab 41 \tab 41 \tab 34 \tab 34 \tab 48 \tab 44 \tab 33 \tab 42 \tab 39 \tab 24 \tab 46 \tab 41 \tab 23 \tab 40 \tab 28 \tab 38 \tab 26 \tab 23 \tab 26 \tab 18 \tab 20 \tab 21 \tab 15 \tab 22 \tab 20 \tab 17 \tab 14 \tab 27 \tab 19 \tab 19 \tab 21 \tab 12 \tab 14 \tab 18 \tab 16 \tab 23 \tab 23 \tab 25 \tab 18 \tab 20 \tab 11 \tab 19 \tab 12 \tab 11 \tab 6 \tab 9 \tab 12 \tab 9 \tab 7 \tab 4 \tab 4 \tab 2 \tab 4 \tab 3 \tab - \tab 13 \tab 6 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 4 \tab 11 \tab 10 \tab 22 \tab 30 \tab 33 \tab 45 \tab 47 \tab 37 \tab 58 \tab 61 \tab 65 \tab 61 \tab 73 \tab 78 \tab 81 \tab 75 \tab 76 \tab 71 \tab 81 \tab 75 \tab 57 \tab 81 \tab 61 \tab 61 \tab 68 \tab 66 \tab 57 \tab 71 \tab 59 \tab 72 \tab 56 \tab 45 \tab 50 \tab 53 \tab 45 \tab 46 \tab 48 \tab 42 \tab 38 \tab 40 \tab 33 \tab 34 \tab 25 \tab 38 \tab 36 \tab 22 \tab 46 \tab 28 \tab 37 \tab 28 \tab 30 \tab 32 \tab 30 \tab 30 \tab 31 \tab 35 \tab 37 \tab 23 \tab 29 \tab 22 \tab 22 \tab 15 \tab 18 \tab 15 \tab 12 \tab 7 \tab 11 \tab 6 \tab 7 \tab 5 \tab 1 \tab 14 \tab 8 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 4 \tab 17 \tab 16 \tab 31 \tab 34 \tab 43 \tab 66 \tab 45 \tab 55 \tab 51 \tab 56 \tab 75 \tab 50 \tab 65 \tab 66 \tab 60 \tab 60 \tab 70 \tab 72 \tab 92 \tab 65 \tab 68 \tab 76 \tab 63 \tab 77 \tab 57 \tab 59 \tab 53 \tab 67 \tab 74 \tab 64 \tab 59 \tab 62 \tab 52 \tab 56 \tab 41 \tab 40 \tab 40 \tab 33 \tab 37 \tab 32 \tab 28 \tab 31 \tab 35 \tab 34 \tab 21 \tab 25 \tab 41 \tab 24 \tab 29 \tab 33 \tab 29 \tab 23 \tab 20 \tab 23 \tab 22 \tab 27 \tab 25 \tab 25 \tab 20 \tab 21 \tab 22 \tab 13 \tab 10 \tab 13 \tab 10 \tab 6 \tab 7 \tab 7 \tab 8 \tab 7 \tab 1 \tab 42 \tab 9 \tab 2 \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 3 \tab 14 \tab 22 \tab 22 \tab 23 \tab 35 \tab 48 \tab 39 \tab 40 \tab 75 \tab 55 \tab 50 \tab 49 \tab 54 \tab 57 \tab 71 \tab 78 \tab 70 \tab 68 \tab 67 \tab 84 \tab 93 \tab 81 \tab 64 \tab 60 \tab 61 \tab 65 \tab 58 \tab 60 \tab 50 \tab 55 \tab 50 \tab 64 \tab 40 \tab 54 \tab 43 \tab 56 \tab 45 \tab 41 \tab 38 \tab 28 \tab 27 \tab 27 \tab 29 \tab 36 \tab 26 \tab 32 \tab 33 \tab 25 \tab 26 \tab 25 \tab 28 \tab 34 \tab 26 \tab 31 \tab 31 \tab 33 \tab 19 \tab 32 \tab 11 \tab 20 \tab 18 \tab 26 \tab 8 \tab 14 \tab 10 \tab 13 \tab 11 \tab 9 \tab 3 \tab 1 \tab 1 \tab 19 \tab 7 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 9 \tab 16 \tab 22 \tab 30 \tab 32 \tab 36 \tab 41 \tab 65 \tab 47 \tab 49 \tab 58 \tab 44 \tab 59 \tab 63 \tab 55 \tab 46 \tab 60 \tab 49 \tab 64 \tab 77 \tab 70 \tab 66 \tab 64 \tab 64 \tab 70 \tab 69 \tab 59 \tab 69 \tab 61 \tab 64 \tab 73 \tab 52 \tab 60 \tab 44 \tab 64 \tab 46 \tab 37 \tab 51 \tab 32 \tab 31 \tab 33 \tab 45 \tab 39 \tab 36 \tab 32 \tab 17 \tab 20 \tab 36 \tab 29 \tab 26 \tab 21 \tab 23 \tab 33 \tab 23 \tab 30 \tab 26 \tab 21 \tab 17 \tab 33 \tab 24 \tab 21 \tab 21 \tab 16 \tab 15 \tab 17 \tab 17 \tab 20 \tab 16 \tab 6 \tab 3 \tab 6 \tab 1 \tab 12 \tab 3 \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 7 \tab 11 \tab 22 \tab 27 \tab 35 \tab 38 \tab 46 \tab 44 \tab 48 \tab 40 \tab 48 \tab 50 \tab 55 \tab 49 \tab 71 \tab 63 \tab 51 \tab 63 \tab 63 \tab 45 \tab 43 \tab 63 \tab 64 \tab 55 \tab 59 \tab 40 \tab 62 \tab 72 \tab 58 \tab 59 \tab 53 \tab 49 \tab 54 \tab 59 \tab 53 \tab 44 \tab 43 \tab 70 \tab 33 \tab 30 \tab 42 \tab 38 \tab 38 \tab 34 \tab 39 \tab 35 \tab 35 \tab 27 \tab 31 \tab 24 \tab 28 \tab 34 \tab 36 \tab 33 \tab 25 \tab 22 \tab 30 \tab 34 \tab 19 \tab 19 \tab 15 \tab 16 \tab 19 \tab 11 \tab 16 \tab 16 \tab 13 \tab 12 \tab 6 \tab 5 \tab 6 \tab 4 \tab 42 \tab 13 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 7 \tab 12 \tab 18 \tab 25 \tab 23 \tab 39 \tab 34 \tab 45 \tab 59 \tab 43 \tab 43 \tab 38 \tab 59 \tab 49 \tab 56 \tab 56 \tab 65 \tab 62 \tab 56 \tab 50 \tab 60 \tab 60 \tab 74 \tab 52 \tab 52 \tab 64 \tab 75 \tab 51 \tab 52 \tab 70 \tab 63 \tab 70 \tab 65 \tab 54 \tab 68 \tab 43 \tab 47 \tab 57 \tab 49 \tab 57 \tab 37 \tab 40 \tab 47 \tab 46 \tab 38 \tab 27 \tab 33 \tab 32 \tab 38 \tab 37 \tab 29 \tab 24 \tab 29 \tab 14 \tab 23 \tab 14 \tab 24 \tab 22 \tab 16 \tab 15 \tab 17 \tab 18 \tab 12 \tab 20 \tab 9 \tab 11 \tab 10 \tab 15 \tab 9 \tab 8 \tab 3 \tab 11 \tab 46 \tab 4 \tab 5 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 8 \tab 13 \tab 25 \tab 36 \tab 34 \tab 47 \tab 63 \tab 76 \tab 68 \tab 75 \tab 66 \tab 64 \tab 67 \tab 68 \tab 76 \tab 78 \tab 98 \tab 100 \tab 98 \tab 86 \tab 100 \tab 72 \tab 105 \tab 101 \tab 95 \tab 88 \tab 94 \tab 99 \tab 113 \tab 96 \tab 95 \tab 101 \tab 122 \tab 66 \tab 95 \tab 98 \tab 94 \tab 85 \tab 82 \tab 64 \tab 83 \tab 78 \tab 79 \tab 47 \tab 66 \tab 66 \tab 50 \tab 53 \tab 47 \tab 40 \tab 48 \tab 34 \tab 38 \tab 44 \tab 43 \tab 40 \tab 39 \tab 31 \tab 27 \tab 31 \tab 31 \tab 30 \tab 25 \tab 26 \tab 28 \tab 15 \tab 18 \tab 17 \tab 8 \tab 9 \tab 8 \tab 82 \tab 80 \tab 15 \tab 8 \tab 4 \tab 3 \tab 6 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 3 \tab 8 \tab 9 \tab 16 \tab 17 \tab 33 \tab 25 \tab 27 \tab 31 \tab 34 \tab 27 \tab 30 \tab 40 \tab 20 \tab 27 \tab 32 \tab 36 \tab 35 \tab 40 \tab 47 \tab 30 \tab 33 \tab 36 \tab 30 \tab 35 \tab 43 \tab 38 \tab 49 \tab 51 \tab 40 \tab 38 \tab 52 \tab 45 \tab 36 \tab 38 \tab 38 \tab 40 \tab 39 \tab 40 \tab 34 \tab 22 \tab 30 \tab 33 \tab 54 \tab 23 \tab 24 \tab 29 \tab 36 \tab 20 \tab 18 \tab 18 \tab 21 \tab 26 \tab 21 \tab 20 \tab 15 \tab 19 \tab 16 \tab 11 \tab 15 \tab 17 \tab 19 \tab 12 \tab 15 \tab 8 \tab 11 \tab 6 \tab 5 \tab 4 \tab 2 \tab 2 \tab - \tab 106 \tab 8 \tab 5 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 3 \tab 9 \tab 14 \tab 19 \tab 16 \tab 27 \tab 29 \tab 39 \tab 38 \tab 34 \tab 33 \tab 34 \tab 48 \tab 32 \tab 34 \tab 27 \tab 35 \tab 37 \tab 33 \tab 34 \tab 29 \tab 32 \tab 44 \tab 41 \tab 37 \tab 46 \tab 35 \tab 40 \tab 43 \tab 34 \tab 30 \tab 47 \tab 48 \tab 33 \tab 42 \tab 41 \tab 31 \tab 51 \tab 35 \tab 23 \tab 36 \tab 37 \tab 37 \tab 39 \tab 34 \tab 37 \tab 23 \tab 26 \tab 31 \tab 35 \tab 27 \tab 24 \tab 22 \tab 17 \tab 13 \tab 18 \tab 18 \tab 17 \tab 10 \tab 12 \tab 16 \tab 10 \tab 12 \tab 11 \tab 13 \tab 14 \tab 8 \tab 9 \tab 7 \tab 6 \tab 7 \tab - \tab 64 \tab 4 \tab 1 \tab 4 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 3 \tab 3 \tab 16 \tab 12 \tab 15 \tab 19 \tab 23 \tab 24 \tab 33 \tab 24 \tab 29 \tab 33 \tab 43 \tab 30 \tab 40 \tab 33 \tab 39 \tab 39 \tab 28 \tab 37 \tab 40 \tab 26 \tab 42 \tab 36 \tab 41 \tab 40 \tab 24 \tab 49 \tab 39 \tab 50 \tab 42 \tab 40 \tab 42 \tab 44 \tab 39 \tab 41 \tab 34 \tab 41 \tab 41 \tab 25 \tab 36 \tab 31 \tab 41 \tab 30 \tab 35 \tab 31 \tab 32 \tab 37 \tab 25 \tab 20 \tab 19 \tab 26 \tab 20 \tab 23 \tab 17 \tab 20 \tab 17 \tab 14 \tab 11 \tab 17 \tab 23 \tab 11 \tab 15 \tab 16 \tab 12 \tab 7 \tab 9 \tab 11 \tab 10 \tab 4 \tab 7 \tab 2 \tab 23 \tab 11 \tab 4 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 1 \tab 4 \tab 7 \tab 11 \tab 16 \tab 21 \tab 31 \tab 28 \tab 32 \tab 29 \tab 38 \tab 50 \tab 44 \tab 46 \tab 38 \tab 40 \tab 37 \tab 38 \tab 48 \tab 47 \tab 45 \tab 37 \tab 53 \tab 43 \tab 54 \tab 48 \tab 27 \tab 38 \tab 42 \tab 50 \tab 39 \tab 44 \tab 66 \tab 44 \tab 74 \tab 43 \tab 51 \tab 71 \tab 59 \tab 34 \tab 47 \tab 62 \tab 41 \tab 45 \tab 40 \tab 29 \tab 41 \tab 36 \tab 37 \tab 39 \tab 22 \tab 36 \tab 31 \tab 32 \tab 22 \tab 23 \tab 18 \tab 22 \tab 24 \tab 23 \tab 24 \tab 21 \tab 16 \tab 13 \tab 11 \tab 6 \tab 12 \tab 13 \tab 12 \tab 4 \tab 7 \tab 83 \tab 55 \tab 6 \tab 3 \tab 5 \tab 3 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2 \tab 8 \tab 15 \tab 23 \tab 31 \tab 34 \tab 30 \tab 36 \tab 48 \tab 35 \tab 60 \tab 48 \tab 41 \tab 31 \tab 44 \tab 43 \tab 61 \tab 52 \tab 41 \tab 47 \tab 43 \tab 47 \tab 43 \tab 32 \tab 41 \tab 56 \tab 46 \tab 41 \tab 37 \tab 52 \tab 45 \tab 45 \tab 72 \tab 57 \tab 72 \tab 61 \tab 49 \tab 56 \tab 64 \tab 65 \tab 60 \tab 70 \tab 48 \tab 48 \tab 45 \tab 45 \tab 47 \tab 48 \tab 43 \tab 51 \tab 49 \tab 33 \tab 36 \tab 25 \tab 40 \tab 26 \tab 21 \tab 29 \tab 28 \tab 23 \tab 20 \tab 24 \tab 17 \tab 18 \tab 15 \tab 22 \tab 11 \tab 7 \tab 7 \tab 4 \tab 7 \tab 50 \tab 73 \tab 13 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 7 \tab 3 \tab 11 \tab 19 \tab 18 \tab 22 \tab 28 \tab 28 \tab 26 \tab 38 \tab 34 \tab 40 \tab 31 \tab 34 \tab 32 \tab 49 \tab 40 \tab 47 \tab 41 \tab 34 \tab 44 \tab 47 \tab 51 \tab 36 \tab 33 \tab 30 \tab 43 \tab 36 \tab 38 \tab 38 \tab 44 \tab 35 \tab 48 \tab 36 \tab 54 \tab 52 \tab 49 \tab 46 \tab 49 \tab 48 \tab 52 \tab 31 \tab 56 \tab 34 \tab 33 \tab 30 \tab 41 \tab 37 \tab 42 \tab 33 \tab 38 \tab 29 \tab 50 \tab 31 \tab 19 \tab 24 \tab 26 \tab 25 \tab 24 \tab 16 \tab 18 \tab 6 \tab 13 \tab 14 \tab 10 \tab 7 \tab 12 \tab 13 \tab 4 \tab 9 \tab 4 \tab 62 \tab 38 \tab 6 \tab 5 \tab 3 \tab 3 \tab 8 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 1779 \tab 6 \tab 4 \tab 16 \tab 13 \tab 17 \tab 15 \tab 20 \tab 20 \tab 26 \tab 35 \tab 29 \tab 26 \tab 26 \tab 28 \tab 38 \tab 35 \tab 29 \tab 28 \tab 36 \tab 17 \tab 33 \tab 42 \tab 31 \tab 26 \tab 33 \tab 27 \tab 28 \tab 31 \tab 22 \tab 31 \tab 21 \tab 18 \tab 26 \tab 27 \tab 23 \tab 23 \tab 24 \tab 23 \tab 31 \tab 20 \tab 26 \tab 28 \tab 47 \tab 29 \tab 44 \tab 28 \tab 26 \tab 38 \tab 35 \tab 20 \tab 46 \tab 34 \tab 36 \tab 24 \tab 23 \tab 13 \tab 26 \tab 22 \tab 24 \tab 15 \tab 13 \tab 16 \tab 10 \tab 8 \tab 12 \tab 4 \tab 10 \tab 10 \tab 5 \tab 3 \tab 4 \tab - \tab 2 \tab 8 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 56 \tab 6 \tab 3544 \cr
   Total \tab 7135 \tab 112 \tab 236 \tab 401 \tab 579 \tab 659 \tab 839 \tab 980 \tab 1035 \tab 1106 \tab 1159 \tab 1172 \tab 1170 \tab 1207 \tab 1172 \tab 1257 \tab 1216 \tab 1321 \tab 1306 \tab 1279 \tab 1306 \tab 1232 \tab 1249 \tab 1370 \tab 1125 \tab 1233 \tab 1208 \tab 1110 \tab 1171 \tab 1154 \tab 1136 \tab 1111 \tab 1107 \tab 1249 \tab 984 \tab 1173 \tab 1023 \tab 979 \tab 1102 \tab 1008 \tab 906 \tab 941 \tab 926 \tab 965 \tab 853 \tab 900 \tab 751 \tab 785 \tab 892 \tab 749 \tab 810 \tab 713 \tab 728 \tab 729 \tab 597 \tab 655 \tab 561 \tab 589 \tab 530 \tab 507 \tab 458 \tab 436 \tab 378 \tab 356 \tab 327 \tab 303 \tab 247 \tab 231 \tab 215 \tab 143 \tab 112 \tab 102 \tab 312 \tab 779 \tab 176 \tab 69 \tab 43 \tab 26 \tab 66 \tab 11 \tab 2 \tab 15 \tab 15 \tab 13 \tab 4032 \tab 1 \tab 1 \tab 1 \tab 56 \tab 6 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{97} 97 or older
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
