% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_582.R
\name{old2}
\alias{old2}
\title{Age of 2nd person}
\source{
General Social Survey https://gss.norc.org
}
\description{
old2
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. c. How old was \link{PERSON} on his/her last birthday?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 6 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 7 \tab 70 \tab 71 \tab 72 \tab 73 \tab 75 \tab 76 \tab 77 \tab 78 \tab 79 \tab 8 \tab 81 \tab 82 \tab 84 \tab 9 \tab 96 \tab don't know \tab no answer \tab 1 \tab 74 \tab 86 \tab 88 \tab 89 \tab 80 \tab 85 \tab 83 \tab 87 \tab 92 \tab 93 \tab 97 or older \tab 90 \tab 95 \tab 91 \tab 94 \tab not available in this release \tab 150 \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 203 \tab 1 \tab 5 \tab 2 \tab 3 \tab 4 \tab 7 \tab 3 \tab 7 \tab 11 \tab 17 \tab 30 \tab 2 \tab 22 \tab 23 \tab 29 \tab 24 \tab 35 \tab 37 \tab 25 \tab 28 \tab 36 \tab 17 \tab 3 \tab 22 \tab 27 \tab 35 \tab 30 \tab 26 \tab 21 \tab 17 \tab 29 \tab 22 \tab 25 \tab 7 \tab 23 \tab 14 \tab 26 \tab 26 \tab 31 \tab 23 \tab 16 \tab 22 \tab 21 \tab 18 \tab 7 \tab 29 \tab 23 \tab 27 \tab 19 \tab 22 \tab 17 \tab 17 \tab 20 \tab 18 \tab 18 \tab 2 \tab 22 \tab 22 \tab 10 \tab 12 \tab 15 \tab 24 \tab 20 \tab 6 \tab 9 \tab 11 \tab 3 \tab 8 \tab 5 \tab 7 \tab 5 \tab 4 \tab 2 \tab 1 \tab 4 \tab 7 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 893 \tab 1 \tab - \tab 5 \tab 4 \tab - \tab 2 \tab 4 \tab 6 \tab 6 \tab 9 \tab 8 \tab 1 \tab 11 \tab 17 \tab 19 \tab 18 \tab 22 \tab 13 \tab 15 \tab 12 \tab 15 \tab 20 \tab 2 \tab 14 \tab 10 \tab 16 \tab 17 \tab 8 \tab 14 \tab 12 \tab 12 \tab 12 \tab 11 \tab 2 \tab 10 \tab 9 \tab 8 \tab 8 \tab 5 \tab 8 \tab 4 \tab 10 \tab 8 \tab 10 \tab 1 \tab 12 \tab 7 \tab 9 \tab 6 \tab 7 \tab 9 \tab 7 \tab 11 \tab 7 \tab 9 \tab 1 \tab 6 \tab 10 \tab 9 \tab 5 \tab 4 \tab 6 \tab 7 \tab 1 \tab 4 \tab 6 \tab 1 \tab 5 \tab 3 \tab 3 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 5 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 256 \tab - \tab 6 \tab 3 \tab 3 \tab 5 \tab 4 \tab 10 \tab 18 \tab 11 \tab 18 \tab 20 \tab 1 \tab 22 \tab 22 \tab 31 \tab 27 \tab 28 \tab 33 \tab 31 \tab 19 \tab 25 \tab 28 \tab 4 \tab 28 \tab 20 \tab 25 \tab 29 \tab 25 \tab 29 \tab 24 \tab 22 \tab 22 \tab 20 \tab 2 \tab 31 \tab 30 \tab 30 \tab 24 \tab 27 \tab 24 \tab 17 \tab 24 \tab 27 \tab 13 \tab 1 \tab 25 \tab 22 \tab 26 \tab 20 \tab 27 \tab 21 \tab 19 \tab 27 \tab 19 \tab 11 \tab 7 \tab 9 \tab 9 \tab 20 \tab 9 \tab 9 \tab 12 \tab 9 \tab 14 \tab 9 \tab 12 \tab 5 \tab 8 \tab 6 \tab 6 \tab 6 \tab 3 \tab 2 \tab 1 \tab 4 \tab 3 \tab 3 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 9 \tab 1 \tab 6 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 295 \tab 1 \tab 1 \tab 5 \tab 7 \tab 5 \tab 6 \tab 7 \tab 8 \tab 12 \tab 25 \tab 15 \tab 2 \tab 26 \tab 25 \tab 35 \tab 37 \tab 32 \tab 33 \tab 34 \tab 33 \tab 36 \tab 32 \tab 2 \tab 30 \tab 38 \tab 34 \tab 20 \tab 27 \tab 15 \tab 24 \tab 20 \tab 23 \tab 20 \tab 3 \tab 22 \tab 25 \tab 19 \tab 14 \tab 16 \tab 17 \tab 12 \tab 24 \tab 17 \tab 22 \tab 4 \tab 20 \tab 17 \tab 24 \tab 18 \tab 21 \tab 13 \tab 14 \tab 21 \tab 21 \tab 6 \tab 6 \tab 21 \tab 12 \tab 13 \tab 11 \tab 14 \tab 7 \tab 7 \tab 10 \tab 4 \tab 9 \tab 2 \tab 6 \tab 3 \tab 14 \tab 7 \tab 5 \tab 4 \tab 2 \tab 4 \tab 3 \tab 6 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 10 \tab 5 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 285 \tab 1 \tab 2 \tab 7 \tab 3 \tab 5 \tab 2 \tab 9 \tab 15 \tab 9 \tab 8 \tab 20 \tab 3 \tab 21 \tab 25 \tab 29 \tab 35 \tab 23 \tab 24 \tab 29 \tab 30 \tab 34 \tab 27 \tab - \tab 36 \tab 32 \tab 19 \tab 23 \tab 26 \tab 25 \tab 17 \tab 18 \tab 16 \tab 19 \tab 7 \tab 19 \tab 15 \tab 16 \tab 21 \tab 23 \tab 21 \tab 17 \tab 16 \tab 16 \tab 21 \tab 1 \tab 22 \tab 20 \tab 15 \tab 15 \tab 15 \tab 22 \tab 19 \tab 19 \tab 13 \tab 19 \tab 5 \tab 19 \tab 10 \tab 17 \tab 11 \tab 14 \tab 14 \tab 9 \tab 9 \tab 16 \tab 9 \tab 7 \tab 7 \tab 5 \tab 10 \tab 7 \tab 4 \tab 2 \tab 4 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 8 \tab - \tab 1 \tab 7 \tab 3 \tab 6 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 412 \tab 1 \tab 8 \tab 7 \tab 14 \tab 13 \tab 5 \tab 10 \tab 12 \tab 19 \tab 19 \tab 18 \tab 5 \tab 23 \tab 31 \tab 30 \tab 45 \tab 36 \tab 39 \tab 29 \tab 37 \tab 45 \tab 40 \tab 8 \tab 32 \tab 25 \tab 35 \tab 35 \tab 27 \tab 29 \tab 23 \tab 23 \tab 28 \tab 19 \tab 4 \tab 14 \tab 20 \tab 13 \tab 17 \tab 24 \tab 16 \tab 27 \tab 16 \tab 23 \tab 19 \tab 6 \tab 30 \tab 27 \tab 27 \tab 23 \tab 22 \tab 23 \tab 29 \tab 10 \tab 23 \tab 15 \tab 5 \tab 26 \tab 13 \tab 19 \tab 13 \tab 15 \tab 20 \tab 10 \tab 11 \tab 11 \tab 9 \tab 9 \tab 7 \tab 6 \tab 7 \tab 6 \tab 5 \tab 3 \tab 4 \tab 4 \tab 1 \tab 7 \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 16 \tab 3 \tab 8 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 311 \tab 2 \tab 8 \tab 5 \tab 8 \tab 10 \tab 6 \tab 4 \tab 11 \tab 11 \tab 21 \tab 7 \tab 3 \tab 13 \tab 21 \tab 18 \tab 25 \tab 38 \tab 39 \tab 43 \tab 41 \tab 33 \tab 36 \tab 3 \tab 28 \tab 34 \tab 35 \tab 25 \tab 21 \tab 40 \tab 26 \tab 20 \tab 20 \tab 12 \tab 3 \tab 22 \tab 18 \tab 19 \tab 25 \tab 12 \tab 25 \tab 25 \tab 16 \tab 13 \tab 19 \tab 2 \tab 28 \tab 20 \tab 18 \tab 18 \tab 15 \tab 25 \tab 18 \tab 29 \tab 20 \tab 23 \tab 7 \tab 12 \tab 18 \tab 18 \tab 7 \tab 16 \tab 7 \tab 13 \tab 6 \tab 12 \tab 12 \tab 2 \tab 8 \tab 4 \tab 6 \tab 6 \tab 5 \tab 2 \tab 2 \tab 4 \tab 1 \tab 10 \tab 1 \tab 2 \tab - \tab 8 \tab - \tab 1 \tab 7 \tab 4 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 327 \tab - \tab 5 \tab 7 \tab 5 \tab 9 \tab 8 \tab 6 \tab 12 \tab 15 \tab 15 \tab 14 \tab 1 \tab 21 \tab 30 \tab 26 \tab 37 \tab 33 \tab 28 \tab 33 \tab 31 \tab 22 \tab 16 \tab 4 \tab 21 \tab 30 \tab 35 \tab 23 \tab 28 \tab 31 \tab 29 \tab 22 \tab 18 \tab 20 \tab 5 \tab 23 \tab 17 \tab 19 \tab 21 \tab 15 \tab 16 \tab 8 \tab 19 \tab 16 \tab 11 \tab 2 \tab 18 \tab 14 \tab 13 \tab 14 \tab 9 \tab 18 \tab 21 \tab 13 \tab 14 \tab 14 \tab 4 \tab 23 \tab 16 \tab 14 \tab 6 \tab 14 \tab 11 \tab 10 \tab 4 \tab 6 \tab 7 \tab 2 \tab 12 \tab 2 \tab 2 \tab 4 \tab 4 \tab 4 \tab 3 \tab 5 \tab 2 \tab 2 \tab 2 \tab 3 \tab 2 \tab 4 \tab - \tab - \tab 7 \tab 4 \tab 4 \tab - \tab - \tab - \tab 4 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 339 \tab 4 \tab 3 \tab 7 \tab 9 \tab 4 \tab 11 \tab 4 \tab 9 \tab 12 \tab 8 \tab 19 \tab 1 \tab 15 \tab 34 \tab 28 \tab 26 \tab 34 \tab 23 \tab 35 \tab 36 \tab 36 \tab 19 \tab 4 \tab 25 \tab 38 \tab 30 \tab 29 \tab 22 \tab 17 \tab 21 \tab 27 \tab 23 \tab 22 \tab 1 \tab 17 \tab 20 \tab 23 \tab 23 \tab 11 \tab 24 \tab 10 \tab 19 \tab 18 \tab 26 \tab 3 \tab 16 \tab 11 \tab 18 \tab 12 \tab 17 \tab 20 \tab 20 \tab 22 \tab 15 \tab 13 \tab 4 \tab 16 \tab 10 \tab 18 \tab 13 \tab 10 \tab 12 \tab 9 \tab 12 \tab 11 \tab 14 \tab 4 \tab 8 \tab 4 \tab 2 \tab 5 \tab 3 \tab 3 \tab 2 \tab - \tab 3 \tab 5 \tab 2 \tab 2 \tab 2 \tab 9 \tab - \tab - \tab 6 \tab 2 \tab 5 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 313 \tab - \tab 4 \tab 1 \tab 2 \tab 11 \tab 6 \tab 7 \tab 5 \tab 10 \tab 13 \tab 13 \tab 4 \tab 12 \tab 27 \tab 23 \tab 30 \tab 30 \tab 31 \tab 36 \tab 25 \tab 25 \tab 30 \tab 6 \tab 27 \tab 30 \tab 30 \tab 19 \tab 28 \tab 24 \tab 21 \tab 31 \tab 28 \tab 30 \tab 4 \tab 22 \tab 33 \tab 19 \tab 15 \tab 17 \tab 20 \tab 14 \tab 18 \tab 16 \tab 13 \tab 1 \tab 22 \tab 12 \tab 9 \tab 13 \tab 12 \tab 11 \tab 10 \tab 13 \tab 20 \tab 13 \tab 2 \tab 18 \tab 12 \tab 17 \tab 18 \tab 5 \tab 14 \tab 9 \tab 11 \tab 7 \tab 10 \tab 3 \tab 4 \tab 8 \tab 6 \tab 6 \tab 5 \tab 3 \tab 4 \tab 5 \tab 3 \tab 8 \tab 5 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 13 \tab - \tab 5 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 424 \tab 3 \tab 9 \tab 6 \tab 9 \tab 4 \tab 10 \tab 12 \tab 12 \tab 16 \tab 25 \tab 27 \tab 1 \tab 24 \tab 34 \tab 24 \tab 26 \tab 30 \tab 22 \tab 34 \tab 41 \tab 36 \tab 39 \tab 5 \tab 27 \tab 21 \tab 35 \tab 42 \tab 37 \tab 35 \tab 31 \tab 25 \tab 34 \tab 29 \tab 5 \tab 31 \tab 20 \tab 18 \tab 32 \tab 22 \tab 22 \tab 24 \tab 19 \tab 16 \tab 11 \tab 4 \tab 20 \tab 19 \tab 22 \tab 18 \tab 14 \tab 14 \tab 11 \tab 28 \tab 13 \tab 18 \tab 4 \tab 15 \tab 12 \tab 13 \tab 11 \tab 15 \tab 17 \tab 13 \tab 8 \tab 7 \tab 9 \tab 5 \tab 6 \tab 8 \tab 6 \tab 5 \tab 5 \tab 3 \tab 1 \tab 6 \tab 3 \tab 8 \tab 4 \tab - \tab 2 \tab 7 \tab - \tab - \tab 13 \tab 3 \tab 6 \tab 3 \tab - \tab 1 \tab 5 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 328 \tab 2 \tab 6 \tab 7 \tab 5 \tab 4 \tab 8 \tab 10 \tab 15 \tab 10 \tab 15 \tab 16 \tab 1 \tab 15 \tab 30 \tab 22 \tab 26 \tab 23 \tab 22 \tab 42 \tab 30 \tab 24 \tab 30 \tab 3 \tab 34 \tab 24 \tab 20 \tab 26 \tab 29 \tab 20 \tab 22 \tab 24 \tab 22 \tab 16 \tab 6 \tab 24 \tab 15 \tab 25 \tab 20 \tab 20 \tab 15 \tab 9 \tab 18 \tab 18 \tab 12 \tab 8 \tab 11 \tab 12 \tab 15 \tab 8 \tab 13 \tab 15 \tab 7 \tab 10 \tab 8 \tab 17 \tab 5 \tab 18 \tab 13 \tab 17 \tab 15 \tab 10 \tab 10 \tab 9 \tab 7 \tab 18 \tab 12 \tab 6 \tab 8 \tab 5 \tab 4 \tab 5 \tab 10 \tab 3 \tab 2 \tab 3 \tab 1 \tab 5 \tab - \tab 3 \tab 1 \tab 11 \tab - \tab - \tab 14 \tab 9 \tab 11 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 327 \tab 3 \tab 2 \tab 3 \tab 7 \tab 7 \tab 5 \tab 8 \tab 9 \tab 13 \tab 13 \tab 17 \tab 2 \tab 21 \tab 27 \tab 22 \tab 21 \tab 22 \tab 23 \tab 23 \tab 31 \tab 30 \tab 24 \tab 6 \tab 30 \tab 24 \tab 36 \tab 26 \tab 24 \tab 26 \tab 21 \tab 23 \tab 25 \tab 25 \tab 5 \tab 26 \tab 22 \tab 31 \tab 16 \tab 24 \tab 24 \tab 26 \tab 20 \tab 21 \tab 16 \tab 7 \tab 25 \tab 12 \tab 16 \tab 8 \tab 22 \tab 14 \tab 9 \tab 11 \tab 15 \tab 12 \tab 3 \tab 12 \tab 14 \tab 12 \tab 8 \tab 11 \tab 21 \tab 18 \tab 14 \tab 8 \tab 6 \tab 5 \tab 6 \tab 8 \tab 4 \tab 11 \tab 7 \tab 2 \tab 3 \tab 5 \tab 2 \tab 4 \tab 4 \tab 3 \tab - \tab 12 \tab - \tab - \tab 11 \tab 1 \tab 8 \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 323 \tab 2 \tab 3 \tab 5 \tab 5 \tab 1 \tab 6 \tab 7 \tab 10 \tab 9 \tab 10 \tab 18 \tab 2 \tab 25 \tab 23 \tab 22 \tab 22 \tab 19 \tab 15 \tab 12 \tab 27 \tab 31 \tab 18 \tab 4 \tab 31 \tab 35 \tab 28 \tab 31 \tab 32 \tab 15 \tab 23 \tab 16 \tab 17 \tab 17 \tab 2 \tab 25 \tab 23 \tab 19 \tab 26 \tab 14 \tab 20 \tab 16 \tab 21 \tab 7 \tab 12 \tab 1 \tab 23 \tab 10 \tab 18 \tab 8 \tab 9 \tab 13 \tab 15 \tab 12 \tab 12 \tab 9 \tab 4 \tab 11 \tab 9 \tab 15 \tab 10 \tab 11 \tab 11 \tab 9 \tab 10 \tab 8 \tab 10 \tab 4 \tab 6 \tab 4 \tab 5 \tab 3 \tab 7 \tab 3 \tab 3 \tab 4 \tab 1 \tab 6 \tab 2 \tab 1 \tab 3 \tab 8 \tab - \tab 2 \tab 10 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 374 \tab - \tab 3 \tab 4 \tab 8 \tab 3 \tab 6 \tab 10 \tab 8 \tab 12 \tab 10 \tab 8 \tab 6 \tab 23 \tab 23 \tab 22 \tab 16 \tab 17 \tab 22 \tab 19 \tab 22 \tab 30 \tab 24 \tab 5 \tab 35 \tab 26 \tab 27 \tab 23 \tab 37 \tab 28 \tab 28 \tab 25 \tab 30 \tab 30 \tab 1 \tab 32 \tab 11 \tab 26 \tab 25 \tab 22 \tab 17 \tab 13 \tab 21 \tab 15 \tab 15 \tab 8 \tab 16 \tab 15 \tab 7 \tab 7 \tab 12 \tab 9 \tab 18 \tab 11 \tab 12 \tab 17 \tab 7 \tab 11 \tab 11 \tab 9 \tab 10 \tab 8 \tab 20 \tab 13 \tab 15 \tab 5 \tab 7 \tab 2 \tab 9 \tab 6 \tab 8 \tab 7 \tab 9 \tab 5 \tab 3 \tab 5 \tab - \tab 4 \tab 3 \tab 2 \tab 1 \tab 12 \tab - \tab 1 \tab 13 \tab 4 \tab 6 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 373 \tab 3 \tab 7 \tab 6 \tab 6 \tab 8 \tab 5 \tab 7 \tab 8 \tab 17 \tab 11 \tab 6 \tab 3 \tab 24 \tab 21 \tab 28 \tab 19 \tab 25 \tab 21 \tab 23 \tab 28 \tab 21 \tab 36 \tab 4 \tab 24 \tab 31 \tab 32 \tab 29 \tab 27 \tab 35 \tab 28 \tab 25 \tab 30 \tab 40 \tab 2 \tab 28 \tab 21 \tab 23 \tab 26 \tab 28 \tab 18 \tab 32 \tab 24 \tab 19 \tab 22 \tab 4 \tab 17 \tab 13 \tab 14 \tab 15 \tab 12 \tab 17 \tab 10 \tab 16 \tab 13 \tab 8 \tab 3 \tab 15 \tab 12 \tab 15 \tab 9 \tab 10 \tab 12 \tab 13 \tab 11 \tab 11 \tab 4 \tab 3 \tab 8 \tab 5 \tab 8 \tab 6 \tab 6 \tab 7 \tab 3 \tab 3 \tab 3 \tab 4 \tab 4 \tab 1 \tab - \tab 8 \tab - \tab - \tab 9 \tab 5 \tab 8 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 752 \tab 4 \tab 12 \tab 12 \tab 18 \tab 19 \tab 20 \tab 19 \tab 15 \tab 14 \tab 16 \tab 13 \tab 5 \tab 16 \tab 27 \tab 28 \tab 41 \tab 57 \tab 45 \tab 36 \tab 59 \tab 36 \tab 52 \tab 12 \tab 66 \tab 55 \tab 55 \tab 61 \tab 60 \tab 52 \tab 48 \tab 62 \tab 47 \tab 43 \tab 17 \tab 46 \tab 45 \tab 50 \tab 56 \tab 39 \tab 41 \tab 38 \tab 52 \tab 50 \tab 32 \tab 12 \tab 36 \tab 38 \tab 32 \tab 18 \tab 18 \tab 29 \tab 20 \tab 22 \tab 29 \tab 19 \tab 10 \tab 18 \tab 15 \tab 27 \tab 22 \tab 23 \tab 11 \tab 13 \tab 26 \tab 17 \tab 10 \tab 14 \tab 23 \tab 11 \tab 9 \tab 18 \tab 6 \tab 13 \tab 5 \tab 5 \tab 6 \tab 9 \tab 5 \tab 3 \tab 3 \tab 10 \tab - \tab - \tab 11 \tab 4 \tab 14 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 743 \tab 2 \tab 14 \tab 15 \tab 14 \tab 14 \tab 19 \tab 12 \tab 19 \tab 22 \tab 25 \tab 36 \tab 13 \tab 25 \tab 31 \tab 27 \tab 55 \tab 37 \tab 54 \tab 42 \tab 40 \tab 55 \tab 36 \tab 22 \tab 49 \tab 47 \tab 53 \tab 45 \tab 50 \tab 55 \tab 51 \tab 50 \tab 52 \tab 40 \tab 6 \tab 53 \tab 40 \tab 42 \tab 48 \tab 37 \tab 45 \tab 34 \tab 41 \tab 53 \tab 34 \tab 12 \tab 39 \tab 30 \tab 28 \tab 36 \tab 24 \tab 22 \tab 22 \tab 18 \tab 22 \tab 21 \tab 15 \tab 16 \tab 11 \tab 14 \tab 20 \tab 17 \tab 15 \tab 14 \tab 13 \tab 15 \tab 14 \tab 11 \tab 11 \tab 12 \tab 8 \tab 8 \tab 12 \tab 7 \tab 6 \tab 5 \tab 4 \tab 6 \tab - \tab 3 \tab 4 \tab 10 \tab - \tab - \tab 32 \tab 12 \tab 5 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 777 \tab 2 \tab 12 \tab 10 \tab 20 \tab 9 \tab 10 \tab 13 \tab 20 \tab 23 \tab 24 \tab 27 \tab 6 \tab 29 \tab 35 \tab 30 \tab 32 \tab 39 \tab 45 \tab 36 \tab 41 \tab 36 \tab 44 \tab 3 \tab 48 \tab 51 \tab 50 \tab 60 \tab 51 \tab 40 \tab 43 \tab 37 \tab 50 \tab 51 \tab 11 \tab 42 \tab 42 \tab 61 \tab 36 \tab 30 \tab 47 \tab 34 \tab 40 \tab 38 \tab 24 \tab 11 \tab 48 \tab 26 \tab 29 \tab 28 \tab 22 \tab 28 \tab 30 \tab 23 \tab 15 \tab 14 \tab 6 \tab 17 \tab 16 \tab 20 \tab 20 \tab 13 \tab 11 \tab 12 \tab 13 \tab 10 \tab 18 \tab 12 \tab 16 \tab 11 \tab 19 \tab 8 \tab 9 \tab 3 \tab 6 \tab 10 \tab 2 \tab 14 \tab 2 \tab 3 \tab 2 \tab 9 \tab - \tab 1 \tab 33 \tab 4 \tab 10 \tab 1 \tab - \tab 2 \tab 5 \tab 3 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 741 \tab 5 \tab 15 \tab 11 \tab 20 \tab 15 \tab 14 \tab 15 \tab 23 \tab 25 \tab 27 \tab 29 \tab 4 \tab 26 \tab 35 \tab 29 \tab 35 \tab 43 \tab 39 \tab 47 \tab 32 \tab 26 \tab 55 \tab 5 \tab 47 \tab 44 \tab 35 \tab 36 \tab 38 \tab 48 \tab 49 \tab 41 \tab 46 \tab 45 \tab 5 \tab 50 \tab 47 \tab 50 \tab 50 \tab 44 \tab 51 \tab 38 \tab 35 \tab 39 \tab 35 \tab 15 \tab 31 \tab 40 \tab 30 \tab 39 \tab 25 \tab 19 \tab 16 \tab 32 \tab 19 \tab 12 \tab 18 \tab 18 \tab 21 \tab 13 \tab 17 \tab 13 \tab 17 \tab 12 \tab 11 \tab 14 \tab 9 \tab 9 \tab 24 \tab 15 \tab 14 \tab 9 \tab 12 \tab 4 \tab 7 \tab 6 \tab 4 \tab 14 \tab 5 \tab 1 \tab 5 \tab 8 \tab - \tab 1 \tab 17 \tab 7 \tab 9 \tab - \tab 3 \tab - \tab 7 \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 883 \tab - \tab 11 \tab 7 \tab 17 \tab 13 \tab 24 \tab 11 \tab 7 \tab 13 \tab 22 \tab 25 \tab 4 \tab 24 \tab 26 \tab 45 \tab 38 \tab 35 \tab 32 \tab 38 \tab 35 \tab 36 \tab 31 \tab 10 \tab 49 \tab 47 \tab 46 \tab 29 \tab 34 \tab 33 \tab 46 \tab 40 \tab 39 \tab 36 \tab 8 \tab 43 \tab 29 \tab 44 \tab 34 \tab 29 \tab 39 \tab 40 \tab 37 \tab 31 \tab 40 \tab 6 \tab 31 \tab 30 \tab 34 \tab 36 \tab 21 \tab 21 \tab 22 \tab 21 \tab 18 \tab 23 \tab 8 \tab 27 \tab 13 \tab 15 \tab 23 \tab 16 \tab 17 \tab 16 \tab 11 \tab 9 \tab 12 \tab 7 \tab 16 \tab 13 \tab 5 \tab 10 \tab 8 \tab 12 \tab 10 \tab 3 \tab 4 \tab 8 \tab 4 \tab 1 \tab 2 \tab 10 \tab - \tab 1 \tab 29 \tab 5 \tab 10 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 707 \tab - \tab 12 \tab 6 \tab 13 \tab 8 \tab 10 \tab 15 \tab 12 \tab 23 \tab 25 \tab 17 \tab 8 \tab 27 \tab 24 \tab 23 \tab 29 \tab 41 \tab 38 \tab 26 \tab 42 \tab 37 \tab 33 \tab 7 \tab 50 \tab 45 \tab 35 \tab 41 \tab 47 \tab 33 \tab 40 \tab 42 \tab 28 \tab 42 \tab 7 \tab 53 \tab 40 \tab 37 \tab 55 \tab 56 \tab 51 \tab 52 \tab 40 \tab 47 \tab 42 \tab 12 \tab 38 \tab 34 \tab 38 \tab 33 \tab 33 \tab 22 \tab 38 \tab 39 \tab 29 \tab 22 \tab 7 \tab 29 \tab 24 \tab 38 \tab 15 \tab 15 \tab 12 \tab 13 \tab 22 \tab 17 \tab 13 \tab 6 \tab 12 \tab 10 \tab 10 \tab 8 \tab 11 \tab 7 \tab 4 \tab 9 \tab 2 \tab 6 \tab 3 \tab 10 \tab 3 \tab 13 \tab - \tab 11 \tab 34 \tab 6 \tab 6 \tab 4 \tab 2 \tab 5 \tab 7 \tab 4 \tab 7 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 1226 \tab 5 \tab 15 \tab 13 \tab 19 \tab 22 \tab 18 \tab 30 \tab 32 \tab 34 \tab 29 \tab 41 \tab 4 \tab 37 \tab 49 \tab 41 \tab 51 \tab 46 \tab 62 \tab 51 \tab 57 \tab 66 \tab 47 \tab 14 \tab 62 \tab 57 \tab 67 \tab 65 \tab 65 \tab 63 \tab 80 \tab 62 \tab 74 \tab 51 \tab 16 \tab 78 \tab 57 \tab 74 \tab 72 \tab 57 \tab 70 \tab 73 \tab 65 \tab 72 \tab 56 \tab 11 \tab 61 \tab 59 \tab 61 \tab 55 \tab 55 \tab 47 \tab 38 \tab 43 \tab 58 \tab 38 \tab 10 \tab 41 \tab 32 \tab 41 \tab 34 \tab 26 \tab 20 \tab 25 \tab 23 \tab 23 \tab 18 \tab 9 \tab 20 \tab 13 \tab 19 \tab 18 \tab 18 \tab 8 \tab 14 \tab 8 \tab 11 \tab 9 \tab 13 \tab 8 \tab 9 \tab 10 \tab - \tab 48 \tab 56 \tab 5 \tab 16 \tab 3 \tab 3 \tab 1 \tab 7 \tab 5 \tab 6 \tab 5 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 527 \tab 4 \tab 2 \tab 10 \tab 12 \tab 14 \tab 8 \tab 10 \tab 9 \tab 20 \tab 18 \tab 13 \tab 2 \tab 22 \tab 24 \tab 14 \tab 23 \tab 12 \tab 22 \tab 24 \tab 29 \tab 19 \tab 24 \tab 7 \tab 31 \tab 13 \tab 20 \tab 27 \tab 30 \tab 21 \tab 38 \tab 21 \tab 27 \tab 31 \tab 7 \tab 34 \tab 19 \tab 25 \tab 31 \tab 32 \tab 30 \tab 22 \tab 32 \tab 18 \tab 30 \tab 10 \tab 38 \tab 24 \tab 30 \tab 17 \tab 39 \tab 26 \tab 25 \tab 19 \tab 20 \tab 16 \tab 7 \tab 23 \tab 20 \tab 19 \tab 10 \tab 15 \tab 18 \tab 14 \tab 11 \tab 10 \tab 10 \tab 3 \tab 13 \tab 4 \tab 5 \tab 9 \tab 6 \tab 7 \tab 5 \tab 10 \tab 4 \tab 12 \tab 5 \tab 4 \tab 2 \tab 9 \tab - \tab 45 \tab 25 \tab 4 \tab 7 \tab 2 \tab 2 \tab 1 \tab 6 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 595 \tab 1 \tab 10 \tab 4 \tab 10 \tab 1 \tab 12 \tab 10 \tab 8 \tab 13 \tab 13 \tab 19 \tab 3 \tab 24 \tab 24 \tab 22 \tab 26 \tab 22 \tab 18 \tab 27 \tab 28 \tab 28 \tab 25 \tab 4 \tab 25 \tab 26 \tab 24 \tab 23 \tab 24 \tab 20 \tab 28 \tab 19 \tab 30 \tab 23 \tab 5 \tab 34 \tab 19 \tab 28 \tab 23 \tab 30 \tab 24 \tab 26 \tab 21 \tab 28 \tab 23 \tab 8 \tab 32 \tab 24 \tab 24 \tab 25 \tab 23 \tab 24 \tab 25 \tab 20 \tab 24 \tab 31 \tab 5 \tab 21 \tab 16 \tab 22 \tab 16 \tab 14 \tab 12 \tab 12 \tab 12 \tab 11 \tab 13 \tab 8 \tab 12 \tab 10 \tab 8 \tab 10 \tab 6 \tab 8 \tab 4 \tab 7 \tab 6 \tab 9 \tab 4 \tab 5 \tab 4 \tab 2 \tab - \tab - \tab 50 \tab 9 \tab 4 \tab 1 \tab 3 \tab - \tab 5 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 533 \tab 2 \tab 5 \tab 8 \tab 13 \tab 8 \tab 8 \tab 4 \tab 8 \tab 14 \tab 11 \tab 12 \tab 6 \tab 21 \tab 24 \tab 20 \tab 20 \tab 21 \tab 23 \tab 25 \tab 29 \tab 20 \tab 26 \tab 5 \tab 39 \tab 21 \tab 33 \tab 29 \tab 23 \tab 28 \tab 29 \tab 21 \tab 25 \tab 26 \tab 7 \tab 28 \tab 22 \tab 38 \tab 14 \tab 27 \tab 27 \tab 24 \tab 28 \tab 30 \tab 30 \tab 4 \tab 28 \tab 27 \tab 37 \tab 24 \tab 17 \tab 21 \tab 32 \tab 15 \tab 20 \tab 16 \tab 6 \tab 20 \tab 22 \tab 17 \tab 17 \tab 17 \tab 15 \tab 10 \tab 16 \tab 6 \tab 18 \tab 3 \tab 18 \tab 7 \tab 10 \tab 5 \tab 12 \tab 8 \tab 3 \tab 6 \tab 3 \tab 5 \tab 4 \tab 4 \tab 3 \tab 2 \tab - \tab 21 \tab 14 \tab 3 \tab 9 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 678 \tab 2 \tab 4 \tab 8 \tab 6 \tab 9 \tab 5 \tab 10 \tab 16 \tab 14 \tab 14 \tab 18 \tab 4 \tab 13 \tab 21 \tab 22 \tab 29 \tab 35 \tab 37 \tab 20 \tab 30 \tab 29 \tab 39 \tab 4 \tab 41 \tab 32 \tab 39 \tab 31 \tab 31 \tab 29 \tab 31 \tab 28 \tab 41 \tab 37 \tab 7 \tab 32 \tab 21 \tab 33 \tab 36 \tab 22 \tab 28 \tab 31 \tab 37 \tab 30 \tab 33 \tab 10 \tab 45 \tab 27 \tab 32 \tab 40 \tab 46 \tab 36 \tab 30 \tab 41 \tab 22 \tab 30 \tab 6 \tab 29 \tab 25 \tab 21 \tab 21 \tab 18 \tab 22 \tab 16 \tab 19 \tab 14 \tab 10 \tab 8 \tab 11 \tab 7 \tab 11 \tab 9 \tab 13 \tab 12 \tab 8 \tab 11 \tab 5 \tab 6 \tab 5 \tab 7 \tab 5 \tab 6 \tab - \tab 39 \tab 68 \tab 2 \tab 12 \tab 4 \tab - \tab - \tab 1 \tab 2 \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 849 \tab 2 \tab 12 \tab 12 \tab 11 \tab 6 \tab 12 \tab 6 \tab 13 \tab 17 \tab 18 \tab 14 \tab 6 \tab 23 \tab 23 \tab 35 \tab 27 \tab 38 \tab 23 \tab 34 \tab 35 \tab 49 \tab 50 \tab 2 \tab 41 \tab 37 \tab 36 \tab 41 \tab 33 \tab 36 \tab 46 \tab 29 \tab 31 \tab 31 \tab 8 \tab 35 \tab 27 \tab 30 \tab 31 \tab 23 \tab 30 \tab 33 \tab 44 \tab 29 \tab 27 \tab 6 \tab 33 \tab 32 \tab 36 \tab 34 \tab 35 \tab 34 \tab 41 \tab 29 \tab 35 \tab 38 \tab 8 \tab 31 \tab 27 \tab 30 \tab 22 \tab 32 \tab 23 \tab 22 \tab 23 \tab 16 \tab 26 \tab 7 \tab 17 \tab 15 \tab 14 \tab 6 \tab 9 \tab 7 \tab 8 \tab 11 \tab 7 \tab 5 \tab 10 \tab 3 \tab 4 \tab 6 \tab - \tab 20 \tab 59 \tab 3 \tab 13 \tab 2 \tab 1 \tab 2 \tab 10 \tab 7 \tab 7 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 677 \tab 3 \tab 6 \tab 3 \tab 4 \tab 6 \tab 4 \tab 8 \tab 17 \tab 17 \tab 16 \tab 16 \tab 5 \tab 13 \tab 25 \tab 27 \tab 21 \tab 33 \tab 30 \tab 21 \tab 27 \tab 29 \tab 30 \tab 3 \tab 31 \tab 37 \tab 23 \tab 24 \tab 33 \tab 33 \tab 24 \tab 34 \tab 40 \tab 22 \tab 2 \tab 44 \tab 20 \tab 29 \tab 30 \tab 25 \tab 33 \tab 23 \tab 32 \tab 23 \tab 27 \tab 5 \tab 41 \tab 23 \tab 36 \tab 29 \tab 24 \tab 30 \tab 34 \tab 29 \tab 26 \tab 25 \tab 1 \tab 32 \tab 21 \tab 19 \tab 24 \tab 14 \tab 23 \tab 16 \tab 12 \tab 23 \tab 14 \tab 5 \tab 21 \tab 19 \tab 10 \tab 9 \tab 10 \tab 3 \tab 7 \tab 8 \tab 6 \tab 7 \tab 1 \tab 6 \tab 4 \tab 5 \tab - \tab 35 \tab 29 \tab - \tab 10 \tab 2 \tab 1 \tab 2 \tab 8 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab 4032 \cr
   2022 \tab 2668 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 5 \tab 3 \tab 2 \tab 11 \tab 11 \tab 19 \tab 18 \tab 9 \tab 20 \tab - \tab 16 \tab 18 \tab 32 \tab 12 \tab 12 \tab 16 \tab 23 \tab 17 \tab 19 \tab 17 \tab 1 \tab 12 \tab 22 \tab 18 \tab 16 \tab 14 \tab 12 \tab 10 \tab 15 \tab 11 \tab 19 \tab - \tab 10 \tab 13 \tab 15 \tab 9 \tab 14 \tab 14 \tab 21 \tab 11 \tab 13 \tab 16 \tab - \tab 23 \tab 11 \tab 23 \tab 9 \tab 15 \tab 17 \tab 9 \tab 20 \tab 19 \tab 15 \tab - \tab 18 \tab 15 \tab 12 \tab 8 \tab 10 \tab 5 \tab 1 \tab 7 \tab 9 \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 9 \tab 1 \tab 2 \tab - \tab 5 \tab 4 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 21 \tab 1 \tab 3544 \cr
   Total \tab 22740 \tab 56 \tab 202 \tab 199 \tab 275 \tab 237 \tab 262 \tab 281 \tab 364 \tab 443 \tab 496 \tab 537 \tab 108 \tab 621 \tab 786 \tab 774 \tab 863 \tab 917 \tab 904 \tab 909 \tab 973 \tab 954 \tab 948 \tab 161 \tab 1039 \tab 975 \tab 1020 \tab 951 \tab 955 \tab 911 \tab 943 \tab 865 \tab 919 \tab 850 \tab 166 \tab 953 \tab 763 \tab 893 \tab 871 \tab 776 \tab 838 \tab 748 \tab 826 \tab 770 \tab 712 \tab 182 \tab 845 \tab 681 \tab 740 \tab 645 \tab 645 \tab 634 \tab 638 \tab 635 \tab 593 \tab 543 \tab 176 \tab 599 \tab 477 \tab 555 \tab 438 \tab 429 \tab 451 \tab 382 \tab 381 \tab 350 \tab 347 \tab 165 \tab 348 \tab 244 \tab 252 \tab 227 \tab 225 \tab 153 \tab 128 \tab 166 \tab 110 \tab 192 \tab 103 \tab 84 \tab 65 \tab 205 \tab 1 \tab 233 \tab 612 \tab 111 \tab 214 \tab 37 \tab 24 \tab 27 \tab 99 \tab 51 \tab 71 \tab 23 \tab 9 \tab 5 \tab 3 \tab 13 \tab 4 \tab 7 \tab 5 \tab 4032 \tab 21 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{97} 97 or older
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
