% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_584.R
\name{old5}
\alias{old5}
\title{Age of 5th person}
\source{
General Social Survey https://gss.norc.org
}
\description{
old5
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. c. How old was \link{PERSON} on his/her last birthday?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 28 \tab 3 \tab 36 \tab 38 \tab 4 \tab 5 \tab 50 \tab 53 \tab 58 \tab 6 \tab 60 \tab 65 \tab 68 \tab 7 \tab 8 \tab 80 \tab 9 \tab 33 \tab 48 \tab 63 \tab don't know \tab 27 \tab 64 \tab 71 \tab 82 \tab 84 \tab no answer \tab 40 \tab 70 \tab 73 \tab 69 \tab 26 \tab 49 \tab 59 \tab 61 \tab 72 \tab 32 \tab 39 \tab 44 \tab 55 \tab 78 \tab 35 \tab 83 \tab 31 \tab 34 \tab 54 \tab 29 \tab 30 \tab 45 \tab 47 \tab 74 \tab 90 \tab 94 \tab 43 \tab 87 \tab 66 \tab 37 \tab 52 \tab 76 \tab 42 \tab 86 \tab 41 \tab 85 \tab 57 \tab 77 \tab 95 \tab 75 \tab not available in this release \tab 51 \tab 56 \tab adult, age unspecified \tab child, age unspecified \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1202 \tab 6 \tab 9 \tab 13 \tab 14 \tab 20 \tab 20 \tab 16 \tab 20 \tab 15 \tab 11 \tab 9 \tab 3 \tab 10 \tab 6 \tab 6 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 11 \tab 1 \tab 1 \tab 11 \tab 9 \tab 1 \tab 1 \tab 1 \tab 12 \tab 1 \tab 1 \tab 1 \tab 12 \tab 20 \tab 1 \tab 19 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1369 \tab 5 \tab 5 \tab 7 \tab 6 \tab 6 \tab 10 \tab 2 \tab 5 \tab 6 \tab 2 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 10 \tab - \tab - \tab 15 \tab 11 \tab - \tab - \tab - \tab 8 \tab - \tab - \tab - \tab 7 \tab 8 \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1264 \tab 5 \tab 13 \tab 21 \tab 14 \tab 19 \tab 20 \tab 16 \tab 15 \tab 14 \tab 11 \tab 6 \tab 8 \tab 9 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 10 \tab - \tab - \tab 9 \tab 13 \tab 1 \tab - \tab - \tab 12 \tab - \tab 1 \tab - \tab 6 \tab 11 \tab - \tab 17 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1293 \tab 6 \tab 14 \tab 14 \tab 15 \tab 22 \tab 15 \tab 16 \tab 6 \tab 12 \tab 12 \tab 5 \tab 3 \tab 11 \tab 1 \tab 3 \tab 3 \tab 2 \tab - \tab - \tab - \tab 7 \tab - \tab - \tab 11 \tab 8 \tab - \tab 1 \tab - \tab 9 \tab - \tab - \tab - \tab 12 \tab 16 \tab - \tab 8 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1283 \tab 9 \tab 4 \tab 10 \tab 6 \tab 15 \tab 4 \tab 10 \tab 5 \tab 11 \tab 7 \tab 9 \tab 3 \tab 10 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 12 \tab - \tab - \tab 9 \tab 13 \tab - \tab - \tab - \tab 7 \tab - \tab - \tab - \tab 9 \tab 10 \tab - \tab 10 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1610 \tab 17 \tab 17 \tab 17 \tab 10 \tab 11 \tab 14 \tab 13 \tab 17 \tab 6 \tab 10 \tab 5 \tab 6 \tab 5 \tab 3 \tab 4 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 15 \tab - \tab - \tab 9 \tab 10 \tab - \tab - \tab - \tab 14 \tab - \tab - \tab - \tab 11 \tab 9 \tab - \tab 10 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1376 \tab 13 \tab 6 \tab 9 \tab 11 \tab 7 \tab 11 \tab 13 \tab 9 \tab 5 \tab 9 \tab 9 \tab 2 \tab 17 \tab 6 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab - \tab 14 \tab - \tab - \tab 14 \tab 9 \tab - \tab - \tab - \tab 12 \tab - \tab - \tab - \tab 11 \tab 10 \tab - \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1280 \tab 10 \tab 7 \tab 2 \tab 11 \tab 8 \tab 15 \tab 9 \tab 6 \tab 9 \tab 3 \tab 2 \tab 5 \tab 14 \tab 1 \tab 3 \tab 5 \tab 2 \tab 3 \tab - \tab 2 \tab 6 \tab - \tab - \tab 14 \tab 11 \tab - \tab - \tab - \tab 14 \tab - \tab 1 \tab - \tab 4 \tab 10 \tab 1 \tab 9 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1347 \tab 10 \tab 11 \tab 6 \tab 5 \tab 10 \tab 11 \tab 14 \tab 17 \tab 5 \tab 5 \tab 4 \tab 3 \tab 9 \tab 1 \tab - \tab - \tab 4 \tab 2 \tab 1 \tab - \tab 10 \tab - \tab - \tab 10 \tab 11 \tab - \tab - \tab - \tab 10 \tab 1 \tab - \tab - \tab 5 \tab 7 \tab - \tab 7 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1265 \tab 8 \tab 14 \tab 8 \tab 6 \tab 5 \tab 9 \tab 9 \tab 9 \tab 6 \tab 5 \tab 7 \tab 5 \tab 15 \tab 1 \tab 3 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab 10 \tab 1 \tab - \tab 8 \tab 10 \tab - \tab - \tab - \tab 16 \tab - \tab - \tab - \tab 11 \tab 12 \tab - \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1577 \tab 14 \tab 15 \tab 4 \tab 13 \tab 9 \tab 16 \tab 7 \tab 11 \tab 9 \tab 4 \tab 8 \tab 5 \tab 16 \tab 2 \tab 4 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab 13 \tab - \tab 1 \tab 11 \tab 15 \tab - \tab - \tab - \tab 15 \tab - \tab - \tab - \tab 7 \tab 9 \tab - \tab 13 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 5 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1328 \tab 13 \tab 6 \tab 8 \tab 7 \tab 6 \tab 3 \tab 3 \tab 4 \tab 4 \tab 5 \tab 2 \tab 2 \tab 5 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 7 \tab - \tab - \tab 5 \tab 9 \tab - \tab - \tab - \tab 14 \tab - \tab - \tab - \tab 13 \tab 9 \tab - \tab 12 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1356 \tab 9 \tab 10 \tab 13 \tab 8 \tab 13 \tab 6 \tab 11 \tab 4 \tab 4 \tab 2 \tab 3 \tab 8 \tab 9 \tab 8 \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab 10 \tab - \tab - \tab 8 \tab 8 \tab - \tab - \tab - \tab 11 \tab - \tab - \tab - \tab 6 \tab 7 \tab - \tab 8 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1237 \tab 6 \tab 9 \tab 11 \tab 2 \tab 3 \tab 4 \tab 5 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 9 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 8 \tab - \tab - \tab 13 \tab 6 \tab - \tab - \tab 1 \tab 14 \tab - \tab - \tab - \tab 8 \tab 5 \tab - \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1367 \tab 8 \tab 9 \tab 5 \tab 8 \tab 8 \tab 9 \tab 3 \tab 5 \tab 4 \tab 1 \tab 5 \tab 3 \tab 12 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab 9 \tab - \tab - \tab 11 \tab 8 \tab - \tab - \tab - \tab 8 \tab - \tab - \tab - \tab 6 \tab 9 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1451 \tab 8 \tab 15 \tab 7 \tab 11 \tab 9 \tab 7 \tab 1 \tab 3 \tab 3 \tab 3 \tab 3 \tab - \tab 9 \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 9 \tab - \tab - \tab 7 \tab 14 \tab - \tab - \tab - \tab 10 \tab - \tab - \tab - \tab 9 \tab 6 \tab - \tab 6 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2729 \tab 12 \tab 16 \tab 10 \tab 6 \tab 13 \tab 9 \tab 11 \tab 10 \tab 4 \tab 7 \tab 2 \tab 5 \tab 21 \tab 3 \tab 4 \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab 15 \tab - \tab - \tab 15 \tab 16 \tab 1 \tab - \tab - \tab 10 \tab - \tab - \tab - \tab 27 \tab 17 \tab - \tab 11 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2637 \tab 11 \tab 19 \tab 14 \tab 13 \tab 12 \tab 6 \tab 8 \tab 8 \tab 9 \tab 4 \tab 4 \tab 8 \tab 14 \tab 2 \tab 3 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 16 \tab - \tab - \tab 13 \tab 18 \tab - \tab - \tab - \tab 15 \tab - \tab - \tab - \tab 19 \tab 20 \tab - \tab 7 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 9 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2588 \tab - \tab 10 \tab 14 \tab 10 \tab 9 \tab 11 \tab 7 \tab 5 \tab 6 \tab 2 \tab 3 \tab 3 \tab 16 \tab 4 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab 17 \tab - \tab 1 \tab 14 \tab 19 \tab 1 \tab - \tab - \tab 13 \tab - \tab - \tab 2 \tab 16 \tab 21 \tab - \tab 14 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 16 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2559 \tab 7 \tab 17 \tab 18 \tab 14 \tab 18 \tab 9 \tab 9 \tab 9 \tab 6 \tab 7 \tab 4 \tab 1 \tab 17 \tab 4 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab 19 \tab - \tab - \tab 11 \tab 15 \tab - \tab - \tab - \tab 10 \tab - \tab - \tab - \tab 16 \tab 14 \tab - \tab 15 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2556 \tab - \tab 22 \tab 7 \tab 17 \tab 6 \tab 3 \tab 5 \tab 4 \tab 6 \tab 6 \tab 4 \tab 4 \tab 19 \tab - \tab 3 \tab 3 \tab - \tab - \tab 1 \tab - \tab 13 \tab - \tab 1 \tab 11 \tab 12 \tab - \tab - \tab - \tab 11 \tab - \tab - \tab - \tab 14 \tab 9 \tab - \tab 12 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 8 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2567 \tab - \tab 18 \tab 17 \tab 12 \tab 11 \tab 11 \tab 10 \tab 11 \tab 8 \tab 1 \tab 2 \tab 5 \tab 12 \tab 3 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab 13 \tab 1 \tab - \tab 15 \tab 11 \tab - \tab - \tab - \tab 12 \tab 2 \tab - \tab - \tab 16 \tab 9 \tab - \tab 18 \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4085 \tab 16 \tab 27 \tab 18 \tab 15 \tab 26 \tab 18 \tab 11 \tab 15 \tab 15 \tab 13 \tab 8 \tab 5 \tab 25 \tab 6 \tab 8 \tab 3 \tab 3 \tab 2 \tab 1 \tab 3 \tab 22 \tab - \tab - \tab 24 \tab 24 \tab 1 \tab - \tab - \tab 28 \tab - \tab - \tab - \tab 21 \tab 21 \tab - \tab 19 \tab - \tab - \tab - \tab 9 \tab - \tab - \tab - \tab - \tab - \tab 6 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1826 \tab 9 \tab 17 \tab 4 \tab 9 \tab 7 \tab 11 \tab 5 \tab 2 \tab 5 \tab 6 \tab 4 \tab 1 \tab 12 \tab 5 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 9 \tab - \tab - \tab 9 \tab 15 \tab - \tab - \tab - \tab 15 \tab - \tab - \tab - \tab 7 \tab 9 \tab - \tab 10 \tab - \tab - \tab - \tab 7 \tab - \tab - \tab - \tab - \tab - \tab 9 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1861 \tab 6 \tab 8 \tab 7 \tab 7 \tab 6 \tab 7 \tab 9 \tab 9 \tab 3 \tab 5 \tab 2 \tab 2 \tab 7 \tab 5 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 3 \tab 9 \tab 1 \tab 1 \tab 6 \tab 11 \tab - \tab - \tab - \tab 10 \tab - \tab - \tab - \tab 15 \tab 7 \tab - \tab 13 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 9 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1775 \tab 12 \tab 8 \tab 6 \tab 10 \tab 14 \tab 2 \tab 7 \tab 7 \tab 4 \tab 7 \tab 4 \tab 1 \tab 12 \tab 2 \tab 3 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab 15 \tab - \tab - \tab 11 \tab 12 \tab - \tab 1 \tab - \tab 5 \tab - \tab 1 \tab - \tab 16 \tab 15 \tab - \tab 8 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2325 \tab 9 \tab 15 \tab 8 \tab 11 \tab 8 \tab 14 \tab 5 \tab 3 \tab 9 \tab 6 \tab 6 \tab 7 \tab 8 \tab 3 \tab 3 \tab 1 \tab 2 \tab 5 \tab 2 \tab - \tab 10 \tab - \tab - \tab 10 \tab 9 \tab 1 \tab - \tab - \tab 12 \tab - \tab - \tab 1 \tab 7 \tab 13 \tab - \tab 8 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 11 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2632 \tab 9 \tab 11 \tab 12 \tab 9 \tab 7 \tab 7 \tab 9 \tab 7 \tab 8 \tab 9 \tab 5 \tab 3 \tab 13 \tab 3 \tab 2 \tab 4 \tab 2 \tab 1 \tab 3 \tab - \tab 12 \tab - \tab - \tab 9 \tab 7 \tab 1 \tab - \tab - \tab 12 \tab - \tab - \tab - \tab 15 \tab 12 \tab - \tab 13 \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab 11 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2192 \tab 3 \tab 14 \tab 8 \tab 11 \tab 6 \tab 5 \tab 2 \tab 1 \tab 7 \tab 5 \tab 3 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 11 \tab - \tab 1 \tab 8 \tab 6 \tab - \tab - \tab - \tab 9 \tab - \tab - \tab - \tab 10 \tab 11 \tab - \tab 5 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3364 \tab 4 \tab 9 \tab 11 \tab 6 \tab 4 \tab 7 \tab 6 \tab 7 \tab 8 \tab 2 \tab 4 \tab 6 \tab 8 \tab 3 \tab 3 \tab 5 \tab 2 \tab 2 \tab - \tab - \tab 6 \tab 1 \tab - \tab 7 \tab 8 \tab 1 \tab - \tab - \tab 15 \tab - \tab - \tab - \tab 9 \tab 5 \tab - \tab 6 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 5 \tab 3544 \cr
   Total \tab 61902 \tab 245 \tab 375 \tab 309 \tab 297 \tab 318 \tab 294 \tab 252 \tab 235 \tab 214 \tab 171 \tab 137 \tab 111 \tab 353 \tab 85 \tab 82 \tab 66 \tab 44 \tab 42 \tab 23 \tab 13 \tab 348 \tab 5 \tab 6 \tab 328 \tab 348 \tab 8 \tab 3 \tab 2 \tab 363 \tab 4 \tab 4 \tab 4 \tab 345 \tab 341 \tab 2 \tab 312 \tab 6 \tab 4 \tab 4 \tab 31 \tab 20 \tab 2 \tab 4 \tab 2 \tab 3 \tab 129 \tab 7 \tab 6 \tab 5 \tab 3 \tab 19 \tab 4 \tab 4 \tab 2 \tab 3 \tab 8 \tab 4 \tab 3 \tab 4 \tab 2 \tab 5 \tab 1 \tab 6 \tab 4 \tab 2 \tab 7 \tab 4 \tab 2 \tab 3 \tab 3 \tab 3 \tab 1 \tab 5 \tab 3 \tab 1 \tab 7 \tab 4 \tab 5 \tab 4 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 4032 \tab 1 \tab 1 \tab 1 \tab 5 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{97} 97 or older
\item \code{98} adult, age unspecified
\item \code{99} child, age unspecified
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
