% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_584.R
\name{old6}
\alias{old6}
\title{old6}
\source{
General Social Survey https://gss.norc.org
}
\description{
Age of 6th person
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. c. How old was \link{PERSON} on his/her last birthday?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 22 \tab 24 \tab 29 \tab 3 \tab 4 \tab 5 \tab 53 \tab 56 \tab 58 \tab 6 \tab 7 \tab 73 \tab 8 \tab 9 \tab 67 \tab no answer \tab 20 \tab 21 \tab 26 \tab 71 \tab 23 \tab 65 \tab 36 \tab 48 \tab 32 \tab 74 \tab 25 \tab 35 \tab 39 \tab 75 \tab 41 \tab 78 \tab 37 \tab 81 \tab 28 \tab 55 \tab 27 \tab 60 \tab 45 \tab 49 \tab 40 \tab 69 \tab 50 \tab 51 \tab 64 \tab 33 \tab 83 \tab don't know \tab 63 \tab 72 \tab 77 \tab 84 \tab 30 \tab 31 \tab 34 \tab 86 \tab 79 \tab 80 \tab not available in this release \tab 151 \tab 52 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1363 \tab 2 \tab 2 \tab 8 \tab 9 \tab 10 \tab 14 \tab 9 \tab 7 \tab 1 \tab 4 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 6 \tab 5 \tab 8 \tab 1 \tab 1 \tab 1 \tab 7 \tab 6 \tab 1 \tab 5 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1452 \tab 2 \tab 3 \tab 4 \tab 1 \tab 2 \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 6 \tab - \tab - \tab - \tab 2 \tab 8 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1413 \tab 4 \tab 7 \tab 3 \tab 8 \tab 11 \tab 6 \tab 6 \tab 3 \tab 4 \tab 4 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 8 \tab 5 \tab 8 \tab - \tab - \tab 1 \tab 5 \tab 6 \tab - \tab 6 \tab 14 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1424 \tab 8 \tab 8 \tab 3 \tab 7 \tab 4 \tab 7 \tab 8 \tab 5 \tab 4 \tab 3 \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab 4 \tab 3 \tab 5 \tab 1 \tab - \tab - \tab 8 \tab 5 \tab - \tab 7 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1392 \tab 5 \tab 4 \tab 7 \tab 3 \tab 6 \tab 8 \tab 5 \tab - \tab 4 \tab 2 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 6 \tab 5 \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1763 \tab 3 \tab 6 \tab 5 \tab 7 \tab 5 \tab 7 \tab 6 \tab 6 \tab 2 \tab 2 \tab 2 \tab - \tab 3 \tab 2 \tab - \tab - \tab 4 \tab 5 \tab 4 \tab - \tab - \tab - \tab 4 \tab 4 \tab - \tab 9 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1513 \tab 2 \tab 3 \tab 4 \tab 2 \tab 6 \tab 4 \tab 2 \tab 4 \tab 3 \tab 2 \tab 3 \tab - \tab 5 \tab - \tab 1 \tab - \tab 7 \tab 5 \tab 6 \tab - \tab - \tab - \tab 4 \tab 5 \tab - \tab 2 \tab 6 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1402 \tab 4 \tab 3 \tab 4 \tab 3 \tab 4 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab 7 \tab 9 \tab 3 \tab - \tab - \tab - \tab 5 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1472 \tab 4 \tab 2 \tab 5 \tab 4 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 6 \tab 5 \tab 5 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1402 \tab 7 \tab 3 \tab 2 \tab 5 \tab 1 \tab 3 \tab 2 \tab 2 \tab 5 \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 7 \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 6 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1727 \tab 7 \tab 10 \tab 4 \tab 2 \tab 3 \tab 6 \tab 2 \tab 3 \tab 3 \tab 2 \tab 1 \tab 4 \tab 4 \tab 1 \tab - \tab - \tab 6 \tab 4 \tab 6 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 6 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1428 \tab 3 \tab 5 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 5 \tab 8 \tab 1 \tab - \tab - \tab - \tab 6 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1464 \tab 2 \tab 4 \tab 4 \tab 2 \tab 2 \tab 2 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 5 \tab 3 \tab 10 \tab - \tab - \tab - \tab 6 \tab 7 \tab - \tab 5 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1325 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab - \tab 7 \tab 3 \tab 3 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1468 \tab 2 \tab 5 \tab 3 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 5 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 4 \tab - \tab - \tab - \tab 4 \tab 3 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1566 \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 5 \tab 3 \tab 2 \tab - \tab - \tab - \tab 5 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2900 \tab 4 \tab 5 \tab 6 \tab 5 \tab 2 \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 9 \tab - \tab - \tab 1 \tab 9 \tab 9 \tab 7 \tab - \tab - \tab - \tab 5 \tab 5 \tab - \tab 5 \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2791 \tab 5 \tab 7 \tab 6 \tab 3 \tab 4 \tab 7 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 10 \tab 2 \tab - \tab - \tab 8 \tab 9 \tab 7 \tab - \tab - \tab - \tab 5 \tab 5 \tab - \tab 4 \tab 10 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2741 \tab - \tab 13 \tab 8 \tab 1 \tab 3 \tab 1 \tab 2 \tab 5 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab 6 \tab 4 \tab 5 \tab - \tab - \tab - \tab 3 \tab 9 \tab 1 \tab 8 \tab 4 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2721 \tab 5 \tab 8 \tab 3 \tab 5 \tab 6 \tab 5 \tab 1 \tab 6 \tab 1 \tab 2 \tab - \tab 3 \tab 6 \tab 1 \tab 1 \tab - \tab 4 \tab 6 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 9 \tab - \tab 7 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2701 \tab - \tab 12 \tab 1 \tab 8 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 5 \tab 6 \tab 4 \tab - \tab - \tab - \tab 4 \tab 5 \tab - \tab 3 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2724 \tab - \tab 5 \tab 6 \tab 3 \tab 7 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 11 \tab - \tab - \tab - \tab 4 \tab 4 \tab 7 \tab 1 \tab - \tab - \tab 6 \tab 4 \tab - \tab 7 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4358 \tab 6 \tab 16 \tab 8 \tab 6 \tab 8 \tab 4 \tab 3 \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab 14 \tab 1 \tab 1 \tab - \tab 13 \tab 9 \tab 13 \tab - \tab - \tab - \tab 2 \tab 12 \tab - \tab 10 \tab 5 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1950 \tab 3 \tab 10 \tab 2 \tab 5 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 7 \tab - \tab - \tab - \tab 3 \tab 6 \tab 1 \tab - \tab - \tab - \tab 6 \tab 3 \tab - \tab 2 \tab 4 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1950 \tab 3 \tab 7 \tab 2 \tab 3 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 6 \tab - \tab 2 \tab 11 \tab - \tab 1 \tab 1 \tab 4 \tab 5 \tab 4 \tab - \tab - \tab - \tab 4 \tab 8 \tab - \tab 4 \tab 9 \tab - \tab 4 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1898 \tab 3 \tab 8 \tab 2 \tab 2 \tab 3 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab 5 \tab 6 \tab 4 \tab - \tab - \tab - \tab 7 \tab 4 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2463 \tab 5 \tab 9 \tab 6 \tab 2 \tab 5 \tab 4 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab 5 \tab 1 \tab - \tab - \tab - \tab 3 \tab 4 \tab - \tab 6 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2770 \tab 3 \tab 10 \tab 2 \tab 2 \tab 2 \tab 4 \tab 4 \tab 5 \tab 2 \tab 2 \tab 2 \tab 1 \tab 5 \tab - \tab 1 \tab 1 \tab 3 \tab 5 \tab 6 \tab - \tab - \tab - \tab 4 \tab 6 \tab - \tab 9 \tab 4 \tab - \tab 6 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2294 \tab - \tab 4 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 6 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 8 \tab - \tab 4 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab 4032 \cr
   2022 \tab 3465 \tab - \tab 5 \tab 2 \tab 5 \tab 6 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab 4 \tab 4 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 3 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 3544 \cr
   Total \tab 65901 \tab 95 \tab 191 \tab 115 \tab 113 \tab 111 \tab 114 \tab 84 \tab 68 \tab 51 \tab 49 \tab 32 \tab 41 \tab 153 \tab 21 \tab 12 \tab 5 \tab 159 \tab 156 \tab 139 \tab 3 \tab 4 \tab 2 \tab 121 \tab 139 \tab 2 \tab 139 \tab 120 \tab 2 \tab 48 \tab 21 \tab 22 \tab 6 \tab 1 \tab 15 \tab 3 \tab 3 \tab 4 \tab 3 \tab 1 \tab 11 \tab 9 \tab 1 \tab 7 \tab 1 \tab 2 \tab 5 \tab 2 \tab 4 \tab 1 \tab 4 \tab 3 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 6 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4032 \tab 2 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{97} 97 or older
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
