% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_34.R
\name{other}
\alias{other}
\title{Other protestant denominations}
\source{
General Social Survey https://gss.norc.org
}
\description{
other
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 7th day adventist \tab advent children \tab assembly of god \tab baptist (northern) \tab brethren church, brethren \tab christ adelphians \tab christian & missionary alliance \tab christian scientist \tab christian; central christian \tab church of christ \tab church of god in christ \tab churches of god (except with christ and holiness) \tab community church \tab congregationalist, 1st congreg \tab covenant \tab disciples of christ \tab dutch reform \tab evangelical reformed \tab evangelical, evangelist \tab first christian \tab four square gospel \tab friends \tab holiness; church of holiness \tab iap \tab jehovah's witnesses \tab lds--reorganized \tab mennonite \tab mind science \tab mormon \tab nazarene \tab other fundamentalist \tab pentecostal \tab quaker \tab reformed \tab reformed united church of christ \tab religious science \tab salvation army \tab unitarian, universalist \tab united church of christ \tab united church, unity church \tab wesleyan \tab apostolic faith \tab christ church unity \tab christian catholic \tab disciples of god \tab don't know \tab evangelist free church \tab first christian disciples of christ \tab full gospel \tab holiness (nazarene) \tab mission covenant \tab pentecostal assembly of god \tab pentecostal church of god \tab pentecostal holiness, holiness pentecostal \tab reformed church of christ \tab united brethren, united brethren in christ \tab united holiness \tab unity \tab wesleyan methodist--pilgrim \tab zion union \tab zion union apostolic--reformed \tab christian reform \tab church of the living god \tab first church \tab no answer \tab swedish mission \tab holiness church of god \tab missionary baptist \tab united church of christianity \tab christian disciples \tab eden evangelist \tab moravian \tab evangelical covenant \tab sanctified, sanctification \tab church of god, saint & christ \tab free will baptist \tab spiritualist \tab witness holiness \tab church of prophecy \tab evangelical congregational \tab ind bible, bible, bible fellowship \tab new testament christian \tab african methodist \tab brethren, plymouth \tab charismatic \tab church of god in christ holiness \tab federated church \tab holy roller \tab house of prayer \tab pentecostal apostolic \tab apostolic christian \tab bible missionary \tab calvary bible \tab christ cathedral of truth \tab christ in christian union \tab church universal and triumphant \tab evangelical methodist \tab free methodist \tab grace brethren \tab latvian lutheran \tab worldwide church of god \tab first reformed \tab mennonite brethren \tab missionary church \tab open bible \tab pilgrim holiness \tab the way ministry \tab zion union apostolic \tab church of christ, evangelical \tab independent \tab the church of god of prophecy \tab chapel of faith \tab christian calvary chapel \tab evangelical united brethren \tab faith gospel tabernacle \tab polish national church \tab church of daniel's band \tab brother of christ \tab church of the first born \tab chinese gospel church \tab independent fundamental church of america \tab primitive baptist \tab grace reformed \tab new age spirituality \tab new song \tab united church of canada \tab christian tabernacle \tab new birth christian \tab american reform \tab apostolic church \tab assyrian evangelist church \tab church of jesus christ of the restoration \tab laotian christian \tab spirit of christ \tab living word \tab zwinglian \tab course in miracles \tab world overcomer outreach ministry \tab spirit filled \tab unity of the brethren \tab amish \tab christian union \tab community christian fellowship \tab community of christ \tab new hope christian fellowship \tab cathedral of joy \tab cornerstone church \tab great faith ministries \tab greater new testament church \tab harvest church \tab life sanctuary \tab rain on us deliverance ministries \tab real life ministries \tab sanctuary \tab shephard's chapel \tab shield of faith ministries \tab the word church \tab united christian \tab vineyard church \tab word of faith church \tab alliance \tab born again \tab journeys \tab national progressive baptist \tab metropolitan community \tab new apostolic \tab empowerment temple \tab faith covenant \tab grace independent baptist church \tab true light church of christ \tab assembly of christ \tab new life \tab pathways christian church \tab the amana church \tab the legacy church \tab armenian apostolic church \tab calvary \tab disciple of jesus \tab ethiopian evangelical church \tab hebrew israelites \tab hebrew roots \tab scandinavian church \tab messianic jew \tab reformed church of jesus christ of latter day saints \tab skipped on web \tab good united \tab pilgrim church \tab Total \cr
   1972 \tab 6 \tab 1 \tab 9 \tab 1 \tab 7 \tab 1 \tab 1 \tab 2 \tab 10 \tab 20 \tab 3 \tab 9 \tab 1 \tab 20 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 3 \tab 1 \tab 1 \tab 8 \tab 1428 \tab 5 \tab 2 \tab 5 \tab 1 \tab 15 \tab 1 \tab 1 \tab 18 \tab 2 \tab 5 \tab 1 \tab 2 \tab 2 \tab 4 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 6 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab 3 \tab 13 \tab 11 \tab 2 \tab 16 \tab - \tab 7 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 10 \tab 1333 \tab 9 \tab 3 \tab 1 \tab - \tab 6 \tab 7 \tab 1 \tab 9 \tab 1 \tab 3 \tab - \tab 1 \tab 3 \tab 2 \tab 10 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 9 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 2 \tab 21 \tab 20 \tab 1 \tab 8 \tab 1 \tab 10 \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 4 \tab 1309 \tab 7 \tab 2 \tab - \tab - \tab 8 \tab 5 \tab - \tab 13 \tab 2 \tab 5 \tab 1 \tab - \tab 1 \tab 3 \tab 12 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 8 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 4 \tab - \tab 3 \tab - \tab 3 \tab - \tab 1 \tab 5 \tab 18 \tab 19 \tab 2 \tab 9 \tab - \tab 12 \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 9 \tab 1305 \tab 8 \tab - \tab - \tab - \tab 10 \tab 5 \tab - \tab 16 \tab - \tab 7 \tab 2 \tab 1 \tab 1 \tab 1 \tab 11 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 12 \tab - \tab 11 \tab - \tab 2 \tab - \tab - \tab 6 \tab 24 \tab 17 \tab 1 \tab 7 \tab - \tab 15 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 2 \tab - \tab - \tab 7 \tab 1304 \tab 13 \tab - \tab 1 \tab - \tab 9 \tab 6 \tab 4 \tab 12 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 2 \tab 12 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 6 \tab - \tab 10 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 15 \tab 22 \tab 1 \tab 16 \tab - \tab 13 \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 5 \tab 1333 \tab 7 \tab 1 \tab - \tab - \tab 11 \tab 8 \tab 2 \tab 20 \tab - \tab 4 \tab - \tab 2 \tab 1 \tab 4 \tab 10 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 7 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 8 \tab - \tab 13 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 17 \tab 21 \tab - \tab 10 \tab 1 \tab 13 \tab - \tab 5 \tab - \tab 1 \tab 4 \tab 3 \tab - \tab 3 \tab 4 \tab 1319 \tab 10 \tab 3 \tab 1 \tab - \tab 10 \tab 7 \tab 1 \tab 9 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 6 \tab 14 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 8 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 5 \tab - \tab - \tab 5 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 4 \tab - \tab 11 \tab - \tab - \tab - \tab 1 \tab 4 \tab 14 \tab 25 \tab 2 \tab 15 \tab 1 \tab 9 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 5 \tab 1270 \tab 8 \tab 2 \tab 1 \tab - \tab 18 \tab 4 \tab - \tab 17 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 4 \tab 2 \tab 11 \tab - \tab 4 \tab - \tab - \tab 7 \tab 21 \tab 15 \tab 6 \tab 12 \tab - \tab 15 \tab - \tab 4 \tab 1 \tab - \tab 2 \tab 5 \tab 3 \tab - \tab 14 \tab 1603 \tab 10 \tab - \tab - \tab - \tab 10 \tab 8 \tab 1 \tab 19 \tab 2 \tab 7 \tab - \tab 1 \tab - \tab 2 \tab 12 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 9 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 13 \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 9 \tab - \tab 9 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 17 \tab 17 \tab 2 \tab 8 \tab - \tab 17 \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 3 \tab 1364 \tab 14 \tab 1 \tab - \tab 1 \tab 32 \tab 7 \tab 3 \tab 23 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 11 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 6 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 7 \tab - \tab 7 \tab - \tab - \tab - \tab - \tab 3 \tab 21 \tab 25 \tab 1 \tab 11 \tab 1 \tab 6 \tab - \tab 2 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab 7 \tab 1241 \tab 9 \tab 3 \tab - \tab - \tab 30 \tab 9 \tab - \tab 15 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 6 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab 14 \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 2 \tab - \tab 8 \tab - \tab 1 \tab - \tab - \tab 2 \tab 26 \tab 18 \tab 1 \tab 10 \tab - \tab 15 \tab - \tab 9 \tab 1 \tab - \tab 3 \tab 3 \tab - \tab - \tab 7 \tab 1313 \tab 7 \tab - \tab - \tab - \tab 30 \tab 9 \tab 1 \tab 19 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 6 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 4 \tab - \tab 8 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 18 \tab 28 \tab 1 \tab 5 \tab - \tab 12 \tab - \tab 2 \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab 3 \tab 1246 \tab 9 \tab - \tab 8 \tab - \tab 36 \tab 10 \tab - \tab 18 \tab - \tab - \tab - \tab - \tab 3 \tab 5 \tab 10 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 7 \tab - \tab 11 \tab - \tab 3 \tab - \tab 1 \tab 3 \tab 26 \tab 22 \tab 7 \tab 13 \tab 3 \tab 11 \tab - \tab 4 \tab - \tab 1 \tab 2 \tab 4 \tab 1 \tab - \tab 10 \tab 1510 \tab 18 \tab 1 \tab - \tab - \tab 30 \tab 7 \tab 2 \tab 31 \tab 4 \tab 6 \tab - \tab 1 \tab - \tab 6 \tab 5 \tab - \tab 1 \tab 8 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 4 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 13 \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 4 \tab - \tab 21 \tab - \tab 2 \tab - \tab - \tab 1 \tab 5 \tab 20 \tab - \tab 9 \tab - \tab 9 \tab 1 \tab 5 \tab 1 \tab - \tab 2 \tab 6 \tab 2 \tab 2 \tab 6 \tab 1261 \tab 8 \tab 1 \tab - \tab 1 \tab 33 \tab 8 \tab 2 \tab 23 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab 6 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 7 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 6 \tab - \tab 11 \tab - \tab 4 \tab - \tab 1 \tab 2 \tab 5 \tab 16 \tab 1 \tab 8 \tab - \tab 6 \tab - \tab 2 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab 6 \tab 1327 \tab 10 \tab - \tab 1 \tab - \tab 30 \tab 6 \tab - \tab 21 \tab 1 \tab 5 \tab - \tab 1 \tab - \tab 4 \tab 13 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 7 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 8 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 7 \tab 1 \tab 13 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 20 \tab 16 \tab 2 \tab 4 \tab - \tab 13 \tab - \tab 8 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 5 \tab 1153 \tab 5 \tab - \tab 1 \tab - \tab 25 \tab 2 \tab 6 \tab 14 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 6 \tab - \tab 1 \tab 8 \tab - \tab - \tab 6 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 4 \tab - \tab 9 \tab - \tab 2 \tab - \tab 1 \tab - \tab 27 \tab 22 \tab 3 \tab 14 \tab 1 \tab 5 \tab - \tab 5 \tab 3 \tab - \tab 3 \tab 6 \tab 1 \tab - \tab 10 \tab 1254 \tab 10 \tab 1 \tab 4 \tab - \tab 32 \tab 10 \tab 3 \tab 27 \tab 2 \tab - \tab - \tab 1 \tab - \tab 7 \tab 8 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab 1 \tab 8 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 7 \tab - \tab 10 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab 18 \tab 26 \tab 2 \tab 12 \tab - \tab 7 \tab - \tab 6 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab - \tab 9 \tab 1379 \tab 13 \tab - \tab 1 \tab - \tab 20 \tab 4 \tab 2 \tab 19 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 4 \tab 4 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab 13 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 13 \tab - \tab 20 \tab - \tab 4 \tab - \tab 3 \tab 5 \tab 25 \tab 35 \tab 4 \tab 15 \tab - \tab 20 \tab 1 \tab 8 \tab 1 \tab - \tab 3 \tab 5 \tab 2 \tab 1 \tab 2 \tab 2643 \tab 22 \tab - \tab 2 \tab - \tab 35 \tab 8 \tab 2 \tab 31 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 14 \tab 6 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 8 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 14 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 13 \tab - \tab 22 \tab 2 \tab 8 \tab - \tab 2 \tab - \tab 11 \tab 23 \tab 1 \tab 12 \tab 1 \tab 13 \tab - \tab 3 \tab 1 \tab 2 \tab 4 \tab 7 \tab 2 \tab - \tab 2 \tab 2596 \tab 16 \tab - \tab 3 \tab - \tab 22 \tab 10 \tab 2 \tab 48 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 8 \tab 9 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 8 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 6 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 13 \tab - \tab 11 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 12 \tab 18 \tab - \tab 18 \tab 1 \tab 12 \tab 1 \tab 4 \tab - \tab - \tab 3 \tab 4 \tab 1 \tab - \tab 3 \tab 2536 \tab 26 \tab 1 \tab 2 \tab - \tab 12 \tab 7 \tab - \tab 41 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 8 \tab 8 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 8 \tab - \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 32 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 12 \tab - \tab 18 \tab - \tab 3 \tab - \tab 2 \tab 2 \tab 11 \tab 31 \tab 5 \tab 11 \tab 1 \tab 16 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab 2514 \tab 17 \tab - \tab 2 \tab - \tab 19 \tab 11 \tab 4 \tab 44 \tab 3 \tab 3 \tab - \tab 2 \tab - \tab 10 \tab 8 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 7 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab 18 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2 \tab - \tab 15 \tab - \tab - \tab 2 \tab - \tab 4 \tab 19 \tab 24 \tab 2 \tab 11 \tab - \tab 11 \tab 1 \tab 3 \tab - \tab - \tab 14 \tab - \tab - \tab - \tab 7 \tab 2410 \tab 26 \tab - \tab 4 \tab - \tab 22 \tab 6 \tab 3 \tab 51 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 8 \tab 9 \tab - \tab 3 \tab 4 \tab - \tab - \tab 1 \tab 14 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab 3 \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab 22 \tab - \tab - \tab 10 \tab - \tab - \tab - \tab - \tab - \tab 13 \tab - \tab 8 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 9 \tab 3 \tab 12 \tab 2 \tab - \tab - \tab - \tab - \tab 26 \tab 20 \tab 2 \tab 6 \tab 1 \tab 11 \tab - \tab 2 \tab - \tab - \tab 5 \tab - \tab 4 \tab - \tab 6 \tab 2448 \tab 18 \tab - \tab 1 \tab 1 \tab 42 \tab 14 \tab 1 \tab 55 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 8 \tab 8 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab 6 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 19 \tab - \tab - \tab 11 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 11 \tab 3 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 17 \tab 3 \tab 25 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 26 \tab 48 \tab 9 \tab 8 \tab 5 \tab 15 \tab 2 \tab 6 \tab 1 \tab - \tab 13 \tab - \tab 5 \tab 1 \tab 14 \tab 3901 \tab 25 \tab 1 \tab 1 \tab - \tab 70 \tab 13 \tab 5 \tab 95 \tab 6 \tab 4 \tab - \tab - \tab 2 \tab 10 \tab 12 \tab - \tab 6 \tab 10 \tab - \tab - \tab - \tab 2 \tab 7 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 12 \tab 1 \tab - \tab 33 \tab - \tab - \tab 9 \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab 22 \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 10 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 19 \tab 18 \tab 1 \tab 9 \tab - \tab 6 \tab - \tab 2 \tab 2 \tab - \tab 11 \tab - \tab 1 \tab - \tab 4 \tab 1725 \tab 19 \tab - \tab 1 \tab - \tab 34 \tab 2 \tab 2 \tab 44 \tab 9 \tab 3 \tab - \tab 2 \tab - \tab 9 \tab 3 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab - \tab 10 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 11 \tab - \tab - \tab 7 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 8 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 7 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab 22 \tab 19 \tab 1 \tab 9 \tab 1 \tab 7 \tab - \tab 4 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 3 \tab 1822 \tab 13 \tab - \tab 3 \tab 1 \tab 15 \tab 4 \tab - \tab 37 \tab 2 \tab 4 \tab - \tab - \tab - \tab 3 \tab 4 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 16 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 6 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 10 \tab - \tab 9 \tab - \tab 1 \tab - \tab 2 \tab - \tab 18 \tab 15 \tab - \tab 12 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab 13 \tab - \tab - \tab - \tab 1 \tab 1740 \tab 14 \tab - \tab 2 \tab - \tab 16 \tab 4 \tab - \tab 40 \tab 5 \tab 3 \tab 1 \tab 1 \tab - \tab 5 \tab 9 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 4 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 8 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 5 \tab - \tab 5 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 13 \tab 17 \tab 2 \tab 1 \tab 1 \tab 7 \tab 1 \tab 4 \tab - \tab - \tab 6 \tab - \tab 1 \tab - \tab 6 \tab 2263 \tab 23 \tab - \tab 1 \tab - \tab 33 \tab 10 \tab - \tab 54 \tab 3 \tab 3 \tab - \tab - \tab 1 \tab 7 \tab 6 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 16 \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 7 \tab 1 \tab 11 \tab - \tab 6 \tab - \tab 1 \tab - \tab 27 \tab 22 \tab 2 \tab 8 \tab - \tab 12 \tab - \tab 3 \tab - \tab - \tab 16 \tab 1 \tab - \tab - \tab 5 \tab 2550 \tab 28 \tab - \tab 2 \tab - \tab 23 \tab 3 \tab - \tab 60 \tab 4 \tab 2 \tab - \tab - \tab - \tab 2 \tab 5 \tab 2 \tab 4 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 19 \tab - \tab - \tab 7 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 6 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 14 \tab - \tab 10 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 21 \tab 13 \tab 3 \tab 5 \tab - \tab 6 \tab 1 \tab 3 \tab - \tab 1 \tab 10 \tab - \tab - \tab - \tab 7 \tab 2080 \tab 13 \tab - \tab 1 \tab - \tab 26 \tab 3 \tab - \tab 43 \tab 2 \tab 4 \tab - \tab - \tab - \tab 5 \tab 8 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 6 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 20 \tab - \tab - \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 11 \tab - \tab 5 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 9 \tab 2 \tab 5 \tab - \tab 6 \tab - \tab 1 \tab - \tab - \tab 9 \tab - \tab 2 \tab - \tab 3 \tab 3734 \tab 22 \tab - \tab 1 \tab - \tab 39 \tab 7 \tab - \tab 34 \tab 2 \tab 6 \tab - \tab - \tab - \tab 8 \tab 9 \tab - \tab 4 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 46 \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 32 \tab - \tab - \tab 4032 \cr
   2022 \tab 11 \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab 15 \tab 3 \tab 4 \tab - \tab 4 \tab - \tab 3 \tab - \tab 1 \tab 12 \tab - \tab - \tab - \tab - \tab 3303 \tab 21 \tab - \tab - \tab - \tab 31 \tab 3 \tab - \tab 37 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 6 \tab 7 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 32 \tab - \tab 1 \tab 5 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 9 \tab 1 \tab 1 \tab 3544 \cr
   Total \tab 270 \tab 13 \tab 362 \tab 5 \tab 70 \tab 5 \tab 34 \tab 73 \tab 586 \tab 707 \tab 75 \tab 330 \tab 21 \tab 367 \tab 14 \tab 112 \tab 27 \tab 11 \tab 167 \tab 75 \tab 33 \tab 13 \tab 198 \tab 63517 \tab 483 \tab 22 \tab 50 \tab 5 \tab 834 \tab 228 \tab 48 \tab 1057 \tab 74 \tab 99 \tab 8 \tab 32 \tab 24 \tab 181 \tab 279 \tab 15 \tab 56 \tab 67 \tab 2 \tab 6 \tab 3 \tab 51 \tab 40 \tab 12 \tab 46 \tab 31 \tab 4 \tab 25 \tab 22 \tab 103 \tab 7 \tab 31 \tab 3 \tab 45 \tab 6 \tab 2 \tab 2 \tab 104 \tab 4 \tab 11 \tab 410 \tab 1 \tab 7 \tab 107 \tab 13 \tab 7 \tab 1 \tab 19 \tab 7 \tab 33 \tab 4 \tab 98 \tab 34 \tab 1 \tab 2 \tab 3 \tab 28 \tab 3 \tab 2 \tab 4 \tab 15 \tab 22 \tab 3 \tab 2 \tab 3 \tab 12 \tab 9 \tab 4 \tab 1 \tab 1 \tab 3 \tab 1 \tab 4 \tab 20 \tab 6 \tab 1 \tab 5 \tab 3 \tab 3 \tab 15 \tab 9 \tab 4 \tab 2 \tab 2 \tab 5 \tab 49 \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 10 \tab 1 \tab 5 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 14 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 5 \tab 11 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 8 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 41 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} hungarian reformed
\item \code{2} evangelical congregational
\item \code{3} ind bible, bible, bible fellowship
\item \code{5} church of prophecy
\item \code{6} new testament christian
\item \code{7} church of god, saint & christ
\item \code{8} moravian
\item \code{9} christian & missionary alliance
\item \code{10} advent children
\item \code{11} spiritualist
\item \code{12} assembly of god
\item \code{13} free methodist
\item \code{14} apostolic faith
\item \code{15} african methodist
\item \code{16} free will baptist
\item \code{17} eden evangelist
\item \code{18} holiness (nazarene)
\item \code{19} baptist (northern)
\item \code{20} brethren church, brethren
\item \code{21} witness holiness
\item \code{22} brethren, plymouth
\item \code{23} united brethren, united brethren in christ
\item \code{24} independent
\item \code{25} christian disciples
\item \code{26} christ in christian union
\item \code{27} open bible
\item \code{28} christian catholic
\item \code{29} christ church unity
\item \code{30} christ adelphians
\item \code{31} christian; central christian
\item \code{32} christian reform
\item \code{33} christian scientist
\item \code{34} church of christ, evangelical
\item \code{35} church of christ
\item \code{36} churches of god (except with christ and holiness)
\item \code{37} church of god in christ
\item \code{38} church of god in christ holiness
\item \code{39} church of the living god
\item \code{40} congregationalist, 1st congreg
\item \code{41} community church
\item \code{42} covenant
\item \code{43} dutch reform
\item \code{44} disciples of christ
\item \code{45} evangelical, evangelist
\item \code{46} evangelical reformed
\item \code{47} evangelist free church
\item \code{48} first church
\item \code{49} first christian disciples of christ
\item \code{50} first reformed
\item \code{51} first christian
\item \code{52} full gospel
\item \code{53} four square gospel
\item \code{54} friends
\item \code{55} holy roller
\item \code{56} holiness; church of holiness
\item \code{57} pilgrim holiness
\item \code{58} jehovah's witnesses
\item \code{59} lds
\item \code{60} lds--mormon
\item \code{61} lds--reorganized
\item \code{62} lds--jesus christ; church of jesus lds
\item \code{63} mennonite
\item \code{64} mormon
\item \code{65} nazarene
\item \code{66} pentecostal assembly of god
\item \code{67} pentecostal church of god
\item \code{68} pentecostal
\item \code{69} pentecostal holiness, holiness pentecostal
\item \code{70} quaker
\item \code{71} reformed
\item \code{72} reformed united church of christ
\item \code{73} reformed church of christ
\item \code{74} religious science
\item \code{75} mind science
\item \code{76} salvation army
\item \code{77} 7th day adventist
\item \code{78} sanctified, sanctification
\item \code{79} united holiness
\item \code{80} unitarian, universalist
\item \code{81} united church of christ
\item \code{82} united church, unity church
\item \code{83} wesleyan
\item \code{84} wesleyan methodist--pilgrim
\item \code{85} zion union
\item \code{86} zion union apostolic
\item \code{87} zion union apostolic--reformed
\item \code{88} disciples of god
\item \code{89} grace reformed
\item \code{90} holiness church of god
\item \code{91} evangelical covenant
\item \code{92} mission covenant
\item \code{93} missionary baptist
\item \code{94} swedish mission
\item \code{95} unity
\item \code{96} united church of christianity
\item \code{97} other fundamentalist
\item \code{98} federated church
\item \code{99} american reform
\item \code{100} grace brethren
\item \code{101} christ in god
\item \code{102} charismatic
\item \code{103} pentecostal apostolic
\item \code{104} house of prayer
\item \code{105} latvian lutheran
\item \code{106} triumph church of god
\item \code{107} apostolic christian
\item \code{108} christ cathedral of truth
\item \code{109} bible missionary
\item \code{110} calvary bible
\item \code{111} amish
\item \code{112} evangelical methodist
\item \code{113} worldwide church of god
\item \code{114} church universal and triumphant
\item \code{115} mennonite brethren
\item \code{116} church of the first born
\item \code{117} missionary church
\item \code{118} the way ministry
\item \code{119} united church of canada
\item \code{120} evangelical united brethren
\item \code{121} the church of god of prophecy
\item \code{122} chapel of faith
\item \code{123} polish national church
\item \code{124} faith gospel tabernacle
\item \code{125} christian calvary chapel
\item \code{126} carmelite
\item \code{127} church of daniel's band
\item \code{128} christian tabernacle
\item \code{129} living word
\item \code{130} true light church of christ
\item \code{131} macedonia
\item \code{132} brother of christ
\item \code{133} primitive baptist
\item \code{134} independent fundamental church of america
\item \code{135} chinese gospel church
\item \code{136} new age spirituality
\item \code{137} new song
\item \code{138} apostolic church
\item \code{139} faith christian
\item \code{140} people's church
\item \code{141} new birth christian
\item \code{142} unity school of christianity
\item \code{143} assyrian evangelist church
\item \code{144} spirit of christ
\item \code{145} church of jesus christ of the restoration
\item \code{146} laotian christian
\item \code{148} schwenkfelder
\item \code{149} polish catholic
\item \code{150} zwinglian
\item \code{151} world overcomer outreach ministry
\item \code{152} course in miracles
\item \code{153} unity of the brethren
\item \code{154} spirit filled
\item \code{155} christian union
\item \code{156} church of living christ
\item \code{157} community of christ
\item \code{158} new hope christian fellowship
\item \code{159} community christian fellowship
\item \code{160} friends in christ
\item \code{161} hawaiian ohana
\item \code{162} reformed church of jesus christ of latter day saints
\item \code{163} swedenurgian/churches of the new jerusalem
\item \code{164} divine science
\item \code{166} united christian
\item \code{167} sanctuary
\item \code{168} rain on us deliverance ministries
\item \code{169} the word church
\item \code{170} cornerstone church
\item \code{171} life sanctuary
\item \code{172} word of faith church
\item \code{173} harvest church
\item \code{174} shephard's chapel
\item \code{175} greater new testament church
\item \code{176} vineyard church
\item \code{177} real life ministries
\item \code{178} cathedral of joy
\item \code{179} great faith ministries
\item \code{180} shield of faith ministries
\item \code{181} born again
\item \code{182} alliance
\item \code{183} jacobite apostolic
\item \code{184} church of god of israel
\item \code{185} journeys
\item \code{186} national progressive baptist
\item \code{187} new apostolic
\item \code{188} metropolitan community
\item \code{189} family life church
\item \code{190} faith fellowship
\item \code{191} faith covenant
\item \code{192} new thought
\item \code{193} free spirit ministry
\item \code{194} jacobite syrian christian church
\item \code{195} the ark church
\item \code{196} empowerment temple
\item \code{197} grace independent baptist church
\item \code{198} new life
\item \code{201} pathways christian church
\item \code{204} universal church of the kingdom of god
\item \code{205} assembly of christ
\item \code{206} the amana church
\item \code{207} the legacy church
\item \code{208} calvary
\item \code{210} ethiopian evangelical church
\item \code{211} disciple of jesus
\item \code{212} scandinavian church
\item \code{213} hebrew roots
\item \code{214} hebrew israelites
\item \code{215} armenian apostolic church
\item \code{216} word of jesus
\item \code{217} old testament
\item \code{218} apostolic church of jesus christ
\item \code{219} messianic jew
\item \code{220} pilgrim church
\item \code{221} good united
\item \code{222} hutterite
\item \code{223} armenian evangelical church
\item \code{224} 3c usa church
\item \code{225} israelite church of god
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
