% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_76.R
\name{owngun}
\alias{owngun}
\title{owngun}
\source{
General Social Survey https://gss.norc.org
}
\description{
Have gun in home
}
\details{
Question 237. Do you happen to have in your home (or garage) any guns or revolvers?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab no answer \tab refused \tab yes \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 773 \tab 9 \tab 15 \tab 707 \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 785 \tab 4 \tab 10 \tab 684 \tab 1 \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 779 \tab 6 \tab 17 \tab 697 \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 748 \tab 7 \tab 2 \tab 771 \tab 2 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 760 \tab 10 \tab 2 \tab 695 \tab 1 \tab - \tab 1468 \cr
   1982 \tab - \tab 1019 \tab 6 \tab 22 \tab 810 \tab 3 \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 795 \tab 6 \tab 9 \tab 662 \tab 1 \tab - \tab 1473 \cr
   1985 \tab - \tab 842 \tab 4 \tab 10 \tab 678 \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 1032 \tab 7 \tab 10 \tab 770 \tab - \tab - \tab 1819 \cr
   1988 \tab 504 \tab 571 \tab 7 \tab 10 \tab 389 \tab - \tab - \tab 1481 \cr
   1989 \tab 504 \tab 555 \tab 3 \tab - \tab 475 \tab - \tab - \tab 1537 \cr
   1990 \tab 455 \tab 520 \tab 9 \tab - \tab 387 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 583 \tab 7 \tab 10 \tab 393 \tab - \tab - \tab 1517 \cr
   1993 \tab 531 \tab 614 \tab 2 \tab 7 \tab 452 \tab - \tab - \tab 1606 \cr
   1994 \tab 996 \tab 1160 \tab 4 \tab 19 \tab 810 \tab 3 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 1142 \tab - \tab 7 \tab 772 \tab 2 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 1215 \tab 6 \tab 6 \tab 654 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 1231 \tab 3 \tab 23 \tab 603 \tab 1 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 605 \tab - \tab 9 \tab 310 \tab - \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 558 \tab 3 \tab 16 \tab 319 \tab 2 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 1307 \tab - \tab 30 \tab 664 \tab 2 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 870 \tab - \tab 15 \tab 464 \tab 3 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 839 \tab - \tab 42 \tab 394 \tab 6 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 841 \tab - \tab 24 \tab 440 \tab 3 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 1128 \tab - \tab 54 \tab 529 \tab 3 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 1246 \tab - \tab 50 \tab 593 \tab 1 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 993 \tab - \tab 39 \tab 537 \tab 5 \tab - \tab 2348 \cr
   2021 \tab - \tab 2529 \tab - \tab 10 \tab 1383 \tab 3 \tab 107 \tab 4032 \cr
   2022 \tab 1203 \tab 1511 \tab - \tab 44 \tab 758 \tab 3 \tab 25 \tab 3544 \cr
   Total \tab 26245 \tab 27551 \tab 103 \tab 512 \tab 17800 \tab 47 \tab 132 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} refused
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
