% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_19.R
\name{padotdat}
\alias{padotdat}
\title{padotdat}
\source{
General Social Survey https://gss.norc.org
}
\description{
Fa. d.o.t. code - relation to data
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 0 \tab 0.02 \tab 0.16 \tab 0.22 \tab 0.23 \tab 0.25 \tab 0.3 \tab 0.35 \tab 0.38 \tab 0.41 \tab 0.61 \tab 0.63 \tab 0.65 \tab 0.89 \tab 0.9 \tab 0.93 \tab 0.99 \tab 1 \tab 1.03 \tab 1.05 \tab 1.06 \tab 1.07 \tab 1.09 \tab 1.1 \tab 1.11 \tab 1.12 \tab 1.15 \tab 1.16 \tab 1.17 \tab 1.19 \tab 1.2 \tab 1.21 \tab 1.22 \tab 1.24 \tab 1.26 \tab 1.27 \tab 1.29 \tab 1.3 \tab 1.33 \tab 1.35 \tab 1.36 \tab 1.39 \tab 1.42 \tab 1.44 \tab 1.45 \tab 1.53 \tab 1.54 \tab 1.55 \tab 1.61 \tab 1.64 \tab 1.68 \tab 1.69 \tab 1.72 \tab 1.74 \tab 1.79 \tab 1.81 \tab 1.83 \tab 1.86 \tab 1.87 \tab 1.9 \tab 1.97 \tab 1.98 \tab 2 \tab 2.09 \tab 2.11 \tab 2.12 \tab 2.15 \tab 2.22 \tab 2.25 \tab 2.29 \tab 2.32 \tab 2.34 \tab 2.35 \tab 2.37 \tab 2.44 \tab 2.46 \tab 2.54 \tab 2.57 \tab 2.62 \tab 2.7 \tab 2.75 \tab 2.83 \tab 2.84 \tab 2.88 \tab 2.89 \tab 2.9 \tab 2.94 \tab 2.95 \tab 2.98 \tab 2.99 \tab 3 \tab 3.03 \tab 3.08 \tab 3.1 \tab 3.11 \tab 3.12 \tab 3.13 \tab 3.27 \tab 3.28 \tab 3.31 \tab 3.37 \tab 3.38 \tab 3.39 \tab 3.43 \tab 3.46 \tab 3.48 \tab 3.52 \tab 3.53 \tab 3.55 \tab 3.58 \tab 3.62 \tab 3.64 \tab 3.65 \tab 3.7 \tab 3.82 \tab 3.91 \tab 3.93 \tab 3.95 \tab 4.01 \tab 4.19 \tab 4.28 \tab 4.33 \tab 4.62 \tab 4.63 \tab 4.68 \tab 4.71 \tab 4.74 \tab 4.75 \tab 4.77 \tab 5 \tab 5.39 \tab 5.44 \tab 5.46 \tab 5.54 \tab 5.55 \tab 5.56 \tab 5.64 \tab 5.67 \tab 5.74 \tab 5.79 \tab 5.85 \tab 5.87 \tab 5.88 \tab 5.94 \tab 6.03 \tab 6.04 \tab 6.18 \tab 6.22 \tab 6.25 \tab 6.26 \tab 6.28 \tab 6.3 \tab 6.31 \tab 6.33 \tab 6.35 \tab 6.37 \tab 6.39 \tab 6.42 \tab 6.43 \tab 6.45 \tab 6.46 \tab 6.49 \tab 6.5 \tab 6.53 \tab 6.58 \tab 6.59 \tab 6.6 \tab 6.61 \tab 6.62 \tab 6.63 \tab 6.64 \tab 6.65 \tab 6.66 \tab 6.69 \tab 6.7 \tab 6.72 \tab 6.73 \tab 6.74 \tab 6.75 \tab 6.8 \tab 6.81 \tab 6.83 \tab 6.84 \tab 6.86 \tab 6.87 \tab 6.89 \tab 6.91 \tab 6.92 \tab 6.93 \tab 6.95 \tab 6.98 \tab 7 \tab iap \tab 0.46 \tab 0.56 \tab 0.75 \tab 1.13 \tab 1.28 \tab 1.63 \tab 1.84 \tab 2.01 \tab 2.06 \tab 2.17 \tab 2.19 \tab 2.42 \tab 2.52 \tab 2.61 \tab 2.8 \tab 2.86 \tab 3.07 \tab 3.3 \tab 3.72 \tab 3.81 \tab 3.88 \tab 4.1 \tab 4.13 \tab 4.52 \tab 5.25 \tab 5.4 \tab 5.81 \tab 5.98 \tab 5.99 \tab 6.05 \tab 6.09 \tab 6.19 \tab 6.24 \tab 6.29 \tab 6.32 \tab 6.36 \tab 6.4 \tab 6.44 \tab 6.55 \tab 6.67 \tab 6.68 \tab 6.82 \tab 0.12 \tab 1.14 \tab 1.4 \tab 1.65 \tab 1.95 \tab 2.08 \tab 2.5 \tab 3.14 \tab 4.17 \tab 4.18 \tab 4.6 \tab 4.85 \tab 5.01 \tab 5.78 \tab 5.83 \tab 5.91 \tab 5.92 \tab 6.2 \tab 6.27 \tab 6.96 \tab 0.96 \tab 1.67 \tab 2.16 \tab 3.22 \tab 3.26 \tab 4.53 \tab 4.67 \tab 5.8 \tab 6.76 \tab 0.82 \tab 1.38 \tab 1.41 \tab 1.51 \tab 1.76 \tab 2.77 \tab 4.22 \tab 6.1 \tab 6.51 \tab 6.78 \tab 6.79 \tab 0.37 \tab 1.85 \tab 2.02 \tab 3.01 \tab 3.06 \tab 5.42 \tab 6.14 \tab 1.32 \tab 1.52 \tab 2.63 \tab 0.34 \tab 1.04 \tab 2.03 \tab 3.15 \tab 3.32 \tab 4.61 \tab 4.25 \tab 4.37 \tab 5.71 \tab 5.6 \tab 5.73 \tab 1.34 \tab 1.23 \tab Total \cr
   1972 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 6 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 5 \tab 10 \tab 21 \tab 3 \tab 4 \tab 1 \tab 1 \tab 3 \tab 9 \tab 2 \tab 1 \tab 13 \tab 2 \tab 4 \tab 28 \tab 2 \tab 3 \tab 4 \tab 2 \tab 5 \tab 3 \tab 4 \tab 2 \tab 2 \tab 2 \tab 9 \tab 9 \tab 1 \tab 3 \tab 3 \tab 2 \tab 3 \tab 1 \tab 307 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 9 \tab 1 \tab 5 \tab 76 \tab 4 \tab 7 \tab 3 \tab 5 \tab 6 \tab 1 \tab 15 \tab 1 \tab 2 \tab 4 \tab 5 \tab 16 \tab 9 \tab 2 \tab 2 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 12 \tab 6 \tab 4 \tab 5 \tab 8 \tab 24 \tab 8 \tab 4 \tab 1 \tab 2 \tab 3 \tab 2 \tab 35 \tab 6 \tab 1 \tab 2 \tab 2 \tab 2 \tab 6 \tab 1 \tab 10 \tab 4 \tab 22 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 4 \tab 8 \tab 1 \tab 6 \tab 2 \tab 7 \tab 1 \tab 6 \tab 1 \tab 2 \tab 1 \tab 4 \tab 5 \tab 1 \tab 1 \tab 3 \tab 2 \tab 9 \tab 25 \tab 1 \tab 14 \tab 7 \tab 7 \tab 2 \tab 6 \tab 1 \tab 2 \tab 5 \tab 1 \tab 3 \tab 7 \tab 1 \tab 2 \tab 2 \tab 1 \tab 13 \tab 1 \tab 1 \tab 13 \tab 4 \tab 2 \tab 6 \tab 1 \tab 4 \tab 2 \tab 13 \tab 12 \tab 1 \tab 41 \tab 2 \tab 2 \tab 4 \tab 4 \tab 1 \tab 8 \tab 3 \tab 1 \tab 4 \tab 1 \tab 6 \tab 7 \tab 5 \tab 3 \tab 5 \tab 2 \tab 29 \tab 40 \tab 2 \tab 2 \tab 3 \tab 3 \tab 2 \tab 39 \tab 276 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1 \tab - \tab - \tab - \tab 5 \tab 3 \tab 5 \tab 1 \tab 2 \tab 5 \tab 3 \tab 2 \tab - \tab 4 \tab - \tab 3 \tab 7 \tab 4 \tab 7 \tab 24 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 10 \tab 6 \tab 4 \tab 13 \tab 3 \tab 2 \tab 23 \tab 3 \tab - \tab 2 \tab 1 \tab 5 \tab 6 \tab 1 \tab 11 \tab 1 \tab 3 \tab 15 \tab 9 \tab 1 \tab 6 \tab 43 \tab 2 \tab 1 \tab 1 \tab 254 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 7 \tab - \tab 3 \tab 84 \tab 1 \tab 6 \tab 4 \tab 3 \tab 8 \tab - \tab 20 \tab - \tab 2 \tab 4 \tab 9 \tab 9 \tab 11 \tab - \tab 1 \tab 1 \tab 4 \tab 6 \tab 2 \tab 1 \tab 4 \tab 7 \tab 3 \tab - \tab 6 \tab 2 \tab 3 \tab 1 \tab 8 \tab 15 \tab 6 \tab 4 \tab 1 \tab 5 \tab 2 \tab 2 \tab 39 \tab 6 \tab 5 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 7 \tab 3 \tab 14 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 6 \tab 3 \tab 1 \tab 1 \tab 3 \tab 6 \tab - \tab 7 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 6 \tab 9 \tab 34 \tab 1 \tab 5 \tab 7 \tab 4 \tab 2 \tab 6 \tab 1 \tab - \tab 5 \tab 2 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 3 \tab 8 \tab - \tab 1 \tab 9 \tab 5 \tab - \tab 7 \tab 2 \tab 10 \tab 4 \tab 6 \tab 13 \tab - \tab 34 \tab - \tab 3 \tab 2 \tab 5 \tab - \tab 12 \tab 5 \tab 1 \tab 3 \tab 3 \tab - \tab 2 \tab 12 \tab 2 \tab 2 \tab - \tab 25 \tab 41 \tab 5 \tab 3 \tab 2 \tab 2 \tab - \tab 43 \tab 191 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1 \tab 2 \tab - \tab - \tab 4 \tab - \tab 2 \tab 2 \tab 6 \tab 3 \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 8 \tab 18 \tab - \tab 2 \tab 4 \tab 1 \tab 7 \tab 2 \tab 7 \tab 1 \tab 9 \tab 3 \tab - \tab 26 \tab 5 \tab 3 \tab 3 \tab 1 \tab 2 \tab 3 \tab 4 \tab 7 \tab - \tab 3 \tab 15 \tab 14 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 264 \tab 1 \tab - \tab 5 \tab 3 \tab - \tab 1 \tab 8 \tab 2 \tab 2 \tab 84 \tab 2 \tab 2 \tab 3 \tab 2 \tab 3 \tab - \tab 10 \tab - \tab 4 \tab 2 \tab 10 \tab 19 \tab 9 \tab 2 \tab 1 \tab 3 \tab 3 \tab 8 \tab 1 \tab 2 \tab 2 \tab 8 \tab 1 \tab 1 \tab 8 \tab 2 \tab 2 \tab 5 \tab 20 \tab 18 \tab 5 \tab 7 \tab 6 \tab 4 \tab 2 \tab 3 \tab 44 \tab 4 \tab 2 \tab - \tab 3 \tab 1 \tab 7 \tab 2 \tab 6 \tab 1 \tab 3 \tab - \tab 6 \tab - \tab - \tab 4 \tab 4 \tab - \tab 1 \tab - \tab 8 \tab 4 \tab 2 \tab 3 \tab 5 \tab 5 \tab 4 \tab 16 \tab - \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab 10 \tab 22 \tab 5 \tab 7 \tab 3 \tab 6 \tab 1 \tab 12 \tab 2 \tab - \tab 4 \tab 4 \tab 4 \tab 8 \tab 2 \tab 2 \tab - \tab - \tab 6 \tab - \tab 1 \tab 3 \tab 6 \tab - \tab 3 \tab 1 \tab 1 \tab 4 \tab 10 \tab 8 \tab 2 \tab 38 \tab 1 \tab 4 \tab - \tab 8 \tab 1 \tab 12 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 3 \tab 5 \tab 1 \tab 5 \tab 2 \tab 16 \tab 31 \tab 9 \tab - \tab 2 \tab 2 \tab - \tab 40 \tab 215 \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab 8 \tab 2 \tab - \tab 6 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab - \tab 5 \tab 5 \tab 9 \tab 11 \tab - \tab 1 \tab 3 \tab - \tab 6 \tab 5 \tab 7 \tab 1 \tab 16 \tab 3 \tab 3 \tab 24 \tab - \tab 4 \tab 2 \tab - \tab 7 \tab 4 \tab 2 \tab 5 \tab - \tab 3 \tab 10 \tab 9 \tab 4 \tab 1 \tab 13 \tab - \tab 2 \tab - \tab 264 \tab 3 \tab - \tab - \tab - \tab 3 \tab 2 \tab 9 \tab - \tab 2 \tab 72 \tab 3 \tab 6 \tab 4 \tab 4 \tab 9 \tab - \tab 26 \tab - \tab 3 \tab 1 \tab 3 \tab 21 \tab 10 \tab 2 \tab - \tab - \tab 2 \tab - \tab 8 \tab 4 \tab 6 \tab 1 \tab - \tab - \tab 11 \tab 3 \tab 9 \tab - \tab 16 \tab 21 \tab 7 \tab 4 \tab 4 \tab - \tab 3 \tab 3 \tab 44 \tab 3 \tab 1 \tab - \tab - \tab - \tab 7 \tab 1 \tab 7 \tab 1 \tab 29 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 5 \tab 4 \tab 8 \tab 2 \tab 24 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 3 \tab 6 \tab 25 \tab 1 \tab 10 \tab 5 \tab 3 \tab 1 \tab 13 \tab 2 \tab 1 \tab 9 \tab 2 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 6 \tab - \tab 1 \tab 2 \tab 5 \tab - \tab 11 \tab - \tab 2 \tab 4 \tab 13 \tab 12 \tab 3 \tab 15 \tab - \tab 4 \tab 1 \tab 14 \tab 2 \tab 19 \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 4 \tab 2 \tab 3 \tab 31 \tab 37 \tab 10 \tab 1 \tab 1 \tab 4 \tab - \tab 41 \tab 190 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab 12 \tab 19 \tab 1 \tab 4 \tab 1 \tab - \tab 5 \tab 7 \tab 5 \tab 1 \tab 21 \tab 3 \tab 5 \tab 33 \tab 3 \tab - \tab 2 \tab - \tab 3 \tab 7 \tab 2 \tab 4 \tab - \tab 4 \tab 11 \tab 8 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 282 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 11 \tab - \tab 5 \tab 63 \tab 6 \tab 6 \tab 6 \tab 4 \tab 9 \tab - \tab 12 \tab - \tab 2 \tab 2 \tab 11 \tab 19 \tab 8 \tab 2 \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 2 \tab 13 \tab 9 \tab 7 \tab 2 \tab 6 \tab - \tab 4 \tab 1 \tab 35 \tab 7 \tab 2 \tab 2 \tab - \tab 2 \tab 5 \tab 1 \tab 7 \tab 2 \tab 28 \tab - \tab 5 \tab - \tab 3 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 6 \tab 2 \tab 4 \tab 5 \tab 19 \tab - \tab 3 \tab 1 \tab - \tab - \tab 3 \tab 4 \tab - \tab 1 \tab 1 \tab 2 \tab 19 \tab 23 \tab 2 \tab 16 \tab 5 \tab 2 \tab 3 \tab 8 \tab 2 \tab - \tab 4 \tab 6 \tab 3 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 9 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 11 \tab 4 \tab 18 \tab 11 \tab 2 \tab 18 \tab - \tab 3 \tab - \tab 15 \tab - \tab 7 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 3 \tab 3 \tab 1 \tab 27 \tab 40 \tab 6 \tab 2 \tab 4 \tab 4 \tab 1 \tab 44 \tab 219 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 3 \tab - \tab - \tab 5 \tab - \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 7 \tab 21 \tab 1 \tab - \tab 4 \tab 1 \tab 6 \tab 5 \tab 3 \tab 4 \tab 13 \tab 5 \tab 1 \tab 35 \tab - \tab 1 \tab 1 \tab 1 \tab 8 \tab 6 \tab 2 \tab 6 \tab 1 \tab 2 \tab 10 \tab 12 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab 254 \tab 2 \tab - \tab 3 \tab - \tab 3 \tab 6 \tab 9 \tab - \tab 1 \tab 63 \tab 1 \tab 5 \tab 4 \tab 4 \tab 13 \tab 2 \tab 15 \tab - \tab 2 \tab 5 \tab 19 \tab 18 \tab 11 \tab - \tab 2 \tab 3 \tab 4 \tab - \tab 6 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 4 \tab 3 \tab 2 \tab 4 \tab 10 \tab 16 \tab 7 \tab 7 \tab 4 \tab - \tab 3 \tab 1 \tab 30 \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab 7 \tab - \tab 7 \tab 1 \tab 21 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 7 \tab 5 \tab 3 \tab 3 \tab 5 \tab 12 \tab - \tab - \tab - \tab - \tab 2 \tab 5 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 11 \tab 44 \tab 4 \tab 14 \tab 5 \tab 2 \tab 6 \tab 13 \tab 4 \tab - \tab 3 \tab 4 \tab 3 \tab 9 \tab 2 \tab 2 \tab - \tab 3 \tab 12 \tab - \tab 2 \tab 2 \tab 4 \tab - \tab 6 \tab 1 \tab 5 \tab 3 \tab 14 \tab 15 \tab 3 \tab 39 \tab - \tab 4 \tab 2 \tab 10 \tab - \tab 21 \tab 4 \tab - \tab 2 \tab 2 \tab 3 \tab 5 \tab 1 \tab 3 \tab 5 \tab - \tab 25 \tab 36 \tab 4 \tab 1 \tab 1 \tab 2 \tab - \tab 38 \tab 236 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 3 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab 3 \tab 9 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab 7 \tab 16 \tab 4 \tab - \tab - \tab - \tab 5 \tab 5 \tab 6 \tab - \tab 5 \tab 2 \tab 7 \tab 37 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 11 \tab 2 \tab 3 \tab 1 \tab 1 \tab 13 \tab 11 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 236 \tab 1 \tab - \tab - \tab - \tab 1 \tab 6 \tab 9 \tab - \tab 5 \tab 74 \tab 4 \tab 4 \tab 2 \tab 12 \tab 18 \tab - \tab 15 \tab - \tab - \tab 2 \tab 17 \tab 11 \tab 10 \tab 3 \tab 3 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab 6 \tab - \tab 2 \tab - \tab 4 \tab 5 \tab 3 \tab 1 \tab 9 \tab 13 \tab 6 \tab 11 \tab 1 \tab - \tab 7 \tab - \tab 54 \tab 3 \tab - \tab 3 \tab 1 \tab 5 \tab 2 \tab 1 \tab 8 \tab - \tab 35 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 3 \tab 3 \tab 8 \tab 5 \tab 20 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 2 \tab 5 \tab 8 \tab 42 \tab 3 \tab 12 \tab 9 \tab 4 \tab 1 \tab 12 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 8 \tab 2 \tab 2 \tab 2 \tab 1 \tab 7 \tab - \tab 3 \tab 1 \tab 6 \tab - \tab 4 \tab - \tab 12 \tab 7 \tab 9 \tab 12 \tab 1 \tab 29 \tab - \tab 3 \tab - \tab 9 \tab 2 \tab 13 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 5 \tab 2 \tab 1 \tab 26 \tab 37 \tab 5 \tab - \tab 5 \tab - \tab 1 \tab 34 \tab 244 \tab - \tab 1 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab - \tab 2 \tab - \tab 4 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 3 \tab 5 \tab 2 \tab 2 \tab - \tab 5 \tab - \tab 4 \tab 2 \tab 6 \tab 8 \tab 18 \tab 8 \tab 1 \tab 1 \tab - \tab 11 \tab 3 \tab 4 \tab 3 \tab 11 \tab 5 \tab 13 \tab 26 \tab 1 \tab 2 \tab 3 \tab 2 \tab 3 \tab 4 \tab 3 \tab 4 \tab 4 \tab 1 \tab 15 \tab 13 \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 4 \tab 228 \tab - \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 10 \tab - \tab 5 \tab 36 \tab 5 \tab 11 \tab 4 \tab 1 \tab 11 \tab 1 \tab 17 \tab - \tab 2 \tab 2 \tab 11 \tab 12 \tab 11 \tab 1 \tab 5 \tab 4 \tab 2 \tab - \tab 1 \tab 4 \tab 4 \tab 1 \tab 3 \tab - \tab 10 \tab 4 \tab 4 \tab 4 \tab 12 \tab 14 \tab 6 \tab 4 \tab 2 \tab - \tab 4 \tab 12 \tab 39 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 3 \tab 8 \tab 1 \tab 26 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 3 \tab - \tab 5 \tab 4 \tab 14 \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 5 \tab 37 \tab - \tab 12 \tab 7 \tab 2 \tab 2 \tab 9 \tab 2 \tab - \tab 5 \tab 2 \tab 2 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 10 \tab - \tab 1 \tab 4 \tab 3 \tab - \tab 6 \tab - \tab 3 \tab 4 \tab 6 \tab 13 \tab 3 \tab 13 \tab - \tab 6 \tab 2 \tab 7 \tab - \tab 12 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab 22 \tab 33 \tab 2 \tab 1 \tab 4 \tab 4 \tab 2 \tab 39 \tab 260 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 17 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1 \tab 1 \tab - \tab - \tab 6 \tab 1 \tab 3 \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 9 \tab - \tab 1 \tab 2 \tab 1 \tab 8 \tab 26 \tab 5 \tab - \tab 5 \tab 6 \tab 4 \tab 3 \tab 4 \tab 2 \tab 5 \tab 4 \tab 11 \tab 28 \tab 1 \tab 3 \tab 2 \tab 1 \tab 6 \tab 8 \tab 2 \tab 4 \tab 4 \tab 1 \tab 6 \tab 12 \tab 3 \tab 17 \tab 3 \tab - \tab - \tab 1 \tab 264 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 5 \tab 11 \tab - \tab 8 \tab 23 \tab 2 \tab 2 \tab 4 \tab 4 \tab 12 \tab 1 \tab 8 \tab - \tab 4 \tab 7 \tab 9 \tab 17 \tab 6 \tab 4 \tab 1 \tab 5 \tab 3 \tab - \tab - \tab 4 \tab 4 \tab 3 \tab 6 \tab - \tab 8 \tab 6 \tab 5 \tab 4 \tab 13 \tab 20 \tab 5 \tab 10 \tab 2 \tab - \tab 3 \tab 9 \tab 48 \tab 6 \tab 4 \tab - \tab - \tab 1 \tab 6 \tab 1 \tab 16 \tab 1 \tab 48 \tab - \tab 4 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 10 \tab 5 \tab 6 \tab 2 \tab 7 \tab 3 \tab 12 \tab - \tab 4 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 2 \tab - \tab - \tab 6 \tab - \tab 40 \tab 2 \tab 17 \tab 13 \tab 6 \tab 4 \tab 8 \tab 3 \tab 1 \tab 8 \tab 3 \tab 4 \tab 5 \tab 2 \tab 3 \tab 2 \tab - \tab 13 \tab 3 \tab 1 \tab 5 \tab 13 \tab - \tab 11 \tab - \tab 12 \tab 1 \tab 17 \tab 11 \tab 4 \tab 38 \tab - \tab 2 \tab 3 \tab 13 \tab 3 \tab 12 \tab 4 \tab 2 \tab 3 \tab 3 \tab 1 \tab 6 \tab 3 \tab 4 \tab 3 \tab 1 \tab 48 \tab 57 \tab 10 \tab - \tab 1 \tab 1 \tab 2 \tab 47 \tab 378 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab 13 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 7 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1 \tab 1 \tab 1 \tab - \tab 6 \tab 3 \tab 10 \tab 3 \tab 4 \tab 6 \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 5 \tab 14 \tab 25 \tab 7 \tab 2 \tab 4 \tab 8 \tab 6 \tab 7 \tab 9 \tab 4 \tab 19 \tab 3 \tab 7 \tab 44 \tab 1 \tab 4 \tab 1 \tab - \tab 12 \tab 9 \tab 3 \tab 5 \tab 1 \tab 4 \tab 17 \tab 14 \tab - \tab 9 \tab 3 \tab - \tab 1 \tab 6 \tab 192 \tab 1 \tab 1 \tab 3 \tab 3 \tab - \tab - \tab 7 \tab - \tab 1 \tab 41 \tab - \tab 14 \tab 3 \tab 2 \tab 16 \tab 1 \tab 19 \tab - \tab - \tab 4 \tab 7 \tab 20 \tab 10 \tab 6 \tab 1 \tab 3 \tab 1 \tab - \tab 10 \tab 3 \tab 7 \tab - \tab 3 \tab 1 \tab 11 \tab 3 \tab 2 \tab 3 \tab 6 \tab 16 \tab 3 \tab 12 \tab - \tab - \tab 3 \tab 12 \tab 49 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab 5 \tab - \tab 14 \tab 4 \tab 41 \tab - \tab 10 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 6 \tab 5 \tab 5 \tab 6 \tab 3 \tab - \tab 16 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab 3 \tab 7 \tab 4 \tab - \tab 1 \tab 5 \tab 6 \tab 12 \tab - \tab 24 \tab 8 \tab 7 \tab 6 \tab 13 \tab - \tab 1 \tab 5 \tab 4 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab 8 \tab - \tab 2 \tab 8 \tab 2 \tab - \tab 7 \tab - \tab 5 \tab 8 \tab 5 \tab 13 \tab 3 \tab 20 \tab 2 \tab 8 \tab 1 \tab 9 \tab - \tab 15 \tab 1 \tab - \tab 2 \tab 2 \tab 7 \tab 6 \tab 3 \tab 1 \tab 1 \tab - \tab 34 \tab 33 \tab 1 \tab 12 \tab 2 \tab 2 \tab - \tab 40 \tab 238 \tab 4 \tab 1 \tab - \tab 4 \tab - \tab - \tab 2 \tab 1 \tab 7 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab 2 \tab 2 \tab 1 \tab 3 \tab 8 \tab 2 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 4 \tab 4 \tab 10 \tab 6 \tab - \tab 2 \tab - \tab 5 \tab - \tab - \tab 3 \tab 28 \tab 5 \tab 7 \tab 52 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 5 \tab 9 \tab 5 \tab - \tab 11 \tab 4 \tab 1 \tab - \tab 231 \tab 3 \tab - \tab 1 \tab - \tab - \tab 6 \tab 12 \tab - \tab 3 \tab 104 \tab 2 \tab 3 \tab 1 \tab 1 \tab 15 \tab - \tab 8 \tab - \tab - \tab 2 \tab 10 \tab 6 \tab 8 \tab - \tab 4 \tab 4 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 6 \tab 3 \tab 1 \tab 5 \tab 3 \tab 19 \tab 3 \tab 6 \tab 1 \tab - \tab 4 \tab 19 \tab 21 \tab 6 \tab 7 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 10 \tab - \tab 32 \tab - \tab 6 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab 7 \tab 4 \tab 3 \tab 1 \tab 2 \tab - \tab 7 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab 5 \tab 2 \tab - \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 9 \tab 5 \tab 31 \tab - \tab 26 \tab - \tab - \tab 4 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 5 \tab 58 \tab 4 \tab - \tab 6 \tab - \tab 3 \tab 3 \tab 3 \tab 9 \tab 1 \tab 24 \tab - \tab 8 \tab 3 \tab 6 \tab 1 \tab 9 \tab - \tab - \tab - \tab 6 \tab 3 \tab 3 \tab 3 \tab 2 \tab - \tab - \tab 33 \tab 36 \tab 3 \tab - \tab 7 \tab 1 \tab - \tab 16 \tab 249 \tab 1 \tab - \tab - \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 3 \tab 1 \tab 1 \tab 1 \tab 6 \tab - \tab 10 \tab 3 \tab 6 \tab 5 \tab 2 \tab 2 \tab 5 \tab 3 \tab 1 \tab - \tab 2 \tab 4 \tab 12 \tab 33 \tab 5 \tab 2 \tab 6 \tab 2 \tab 6 \tab 6 \tab 7 \tab 6 \tab 20 \tab 4 \tab 6 \tab 35 \tab 5 \tab 3 \tab 1 \tab - \tab 10 \tab 9 \tab 2 \tab 6 \tab 6 \tab 2 \tab 12 \tab 11 \tab 2 \tab 7 \tab 3 \tab - \tab - \tab 4 \tab 227 \tab 1 \tab - \tab 2 \tab 5 \tab 1 \tab - \tab 14 \tab - \tab 5 \tab 56 \tab 1 \tab 8 \tab 10 \tab 8 \tab 16 \tab - \tab 14 \tab - \tab 3 \tab 3 \tab 12 \tab 22 \tab 12 \tab 3 \tab 2 \tab 4 \tab 1 \tab - \tab 5 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 3 \tab 3 \tab 2 \tab 13 \tab 14 \tab 11 \tab 8 \tab 1 \tab - \tab 6 \tab 6 \tab 32 \tab 2 \tab 2 \tab - \tab - \tab - \tab 9 \tab - \tab 17 \tab 2 \tab 29 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 5 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 3 \tab 3 \tab - \tab - \tab 7 \tab 2 \tab 14 \tab - \tab 12 \tab 8 \tab 2 \tab 3 \tab 1 \tab 4 \tab - \tab 6 \tab - \tab 2 \tab 10 \tab - \tab 1 \tab - \tab 1 \tab 6 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab 4 \tab 1 \tab 14 \tab 2 \tab 3 \tab 6 \tab 3 \tab 23 \tab - \tab 2 \tab 4 \tab 4 \tab - \tab 16 \tab 8 \tab 1 \tab 1 \tab 1 \tab 8 \tab 2 \tab 6 \tab 2 \tab 2 \tab 1 \tab 38 \tab 42 \tab - \tab 1 \tab 4 \tab 3 \tab 1 \tab 35 \tab 219 \tab 4 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 3 \tab 5 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 3 \tab 2 \tab - \tab - \tab 4 \tab 1 \tab 4 \tab 2 \tab 2 \tab 8 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 2 \tab 5 \tab 8 \tab 17 \tab - \tab 6 \tab 3 \tab 3 \tab 10 \tab 7 \tab 7 \tab 4 \tab 19 \tab 3 \tab 6 \tab 41 \tab 4 \tab 5 \tab - \tab - \tab 3 \tab 6 \tab 2 \tab 6 \tab 2 \tab 4 \tab 5 \tab 5 \tab 2 \tab 2 \tab - \tab - \tab - \tab 7 \tab 214 \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab 3 \tab 6 \tab - \tab 3 \tab 53 \tab 1 \tab 7 \tab 2 \tab 5 \tab 15 \tab 1 \tab 26 \tab - \tab - \tab 3 \tab 7 \tab 7 \tab 12 \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab 5 \tab 2 \tab 5 \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab 4 \tab 3 \tab 13 \tab 12 \tab 9 \tab 6 \tab - \tab - \tab 2 \tab 6 \tab 33 \tab - \tab 4 \tab - \tab 2 \tab 2 \tab 3 \tab - \tab 3 \tab - \tab 55 \tab - \tab 4 \tab - \tab 1 \tab - \tab 4 \tab - \tab 4 \tab - \tab - \tab 4 \tab 7 \tab 6 \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 10 \tab 2 \tab 11 \tab 5 \tab 3 \tab 3 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 2 \tab 12 \tab - \tab 1 \tab 5 \tab 4 \tab 1 \tab 5 \tab - \tab 6 \tab 3 \tab 4 \tab 11 \tab - \tab 24 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 19 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 4 \tab 4 \tab 4 \tab 1 \tab 3 \tab 32 \tab 20 \tab 1 \tab 7 \tab 6 \tab 2 \tab 1 \tab 40 \tab 258 \tab 5 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 18 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 8 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 6 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab 3 \tab 7 \tab 1 \tab 4 \tab 11 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab 15 \tab 24 \tab 1 \tab 1 \tab 2 \tab 9 \tab 2 \tab 6 \tab 9 \tab 2 \tab 11 \tab 2 \tab 9 \tab 40 \tab 3 \tab 2 \tab 1 \tab - \tab 7 \tab 10 \tab 1 \tab 1 \tab 3 \tab 3 \tab 20 \tab 7 \tab 2 \tab 10 \tab 3 \tab 6 \tab 2 \tab 4 \tab 271 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 10 \tab - \tab 11 \tab 34 \tab 9 \tab 12 \tab 4 \tab 5 \tab 14 \tab - \tab 11 \tab - \tab 3 \tab 2 \tab 6 \tab 13 \tab 9 \tab 2 \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab - \tab 10 \tab 18 \tab 2 \tab 6 \tab 2 \tab - \tab 2 \tab 12 \tab 44 \tab 9 \tab 5 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 11 \tab - \tab 35 \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 4 \tab - \tab 4 \tab 5 \tab 4 \tab 1 \tab 2 \tab 2 \tab 20 \tab 1 \tab 5 \tab - \tab - \tab - \tab 6 \tab 5 \tab - \tab - \tab - \tab 6 \tab 3 \tab 26 \tab 2 \tab 14 \tab 10 \tab 11 \tab 8 \tab 4 \tab 2 \tab 1 \tab 4 \tab 1 \tab 3 \tab 3 \tab 2 \tab 4 \tab 3 \tab 5 \tab 8 \tab - \tab - \tab 7 \tab 3 \tab 1 \tab 6 \tab - \tab 9 \tab 5 \tab 5 \tab 11 \tab 3 \tab 39 \tab - \tab 1 \tab 8 \tab 9 \tab - \tab 20 \tab 2 \tab 5 \tab 2 \tab 2 \tab 2 \tab 1 \tab 8 \tab 2 \tab 6 \tab 1 \tab 41 \tab 65 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab 52 \tab 341 \tab 6 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab 8 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab 3 \tab - \tab - \tab - \tab 6 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1819 \cr
   1988 \tab 2 \tab - \tab - \tab - \tab 6 \tab 5 \tab 7 \tab 2 \tab 3 \tab 7 \tab 3 \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 10 \tab 16 \tab 6 \tab 5 \tab 6 \tab 1 \tab 4 \tab 6 \tab 6 \tab 5 \tab 8 \tab 10 \tab 9 \tab 41 \tab 1 \tab 4 \tab 1 \tab - \tab 4 \tab 5 \tab - \tab 7 \tab 4 \tab 5 \tab 11 \tab 12 \tab 2 \tab 4 \tab 5 \tab 1 \tab 1 \tab 2 \tab 202 \tab 2 \tab - \tab - \tab 2 \tab 8 \tab 6 \tab 8 \tab - \tab 4 \tab 58 \tab 3 \tab 12 \tab 3 \tab 7 \tab 12 \tab - \tab 25 \tab - \tab 1 \tab 3 \tab 3 \tab 18 \tab 9 \tab 3 \tab 3 \tab 3 \tab - \tab - \tab 2 \tab 6 \tab 5 \tab - \tab 1 \tab - \tab 4 \tab 6 \tab 7 \tab 3 \tab 8 \tab 7 \tab 2 \tab 8 \tab 3 \tab - \tab 3 \tab 2 \tab 29 \tab 7 \tab 4 \tab 2 \tab - \tab - \tab 1 \tab - \tab 6 \tab 1 \tab 44 \tab - \tab 5 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 8 \tab 5 \tab 3 \tab 2 \tab - \tab 2 \tab 8 \tab - \tab 9 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 4 \tab 4 \tab 12 \tab 3 \tab 3 \tab 7 \tab 3 \tab 2 \tab 4 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 5 \tab 2 \tab 1 \tab - \tab 2 \tab 14 \tab - \tab 3 \tab 7 \tab 8 \tab - \tab 8 \tab - \tab - \tab 3 \tab 4 \tab 7 \tab 2 \tab 18 \tab - \tab 6 \tab 2 \tab 8 \tab 2 \tab 13 \tab 1 \tab 2 \tab 3 \tab 3 \tab 5 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 35 \tab 32 \tab - \tab 2 \tab 5 \tab - \tab - \tab 42 \tab 242 \tab 8 \tab - \tab 2 \tab 4 \tab - \tab - \tab 2 \tab 2 \tab 12 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 7 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1481 \cr
   1989 \tab 1 \tab 2 \tab 1 \tab 2 \tab 9 \tab 2 \tab 6 \tab 4 \tab 3 \tab 7 \tab 2 \tab 1 \tab 3 \tab 3 \tab - \tab 3 \tab 3 \tab 4 \tab 7 \tab 22 \tab 4 \tab 4 \tab 6 \tab 5 \tab 6 \tab 4 \tab 2 \tab 2 \tab 14 \tab 7 \tab 8 \tab 49 \tab 2 \tab 3 \tab 3 \tab - \tab 7 \tab 7 \tab 1 \tab 5 \tab 1 \tab 6 \tab 12 \tab 10 \tab 1 \tab 6 \tab 4 \tab 2 \tab - \tab 5 \tab 200 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab 5 \tab - \tab 8 \tab 71 \tab 6 \tab 11 \tab 7 \tab 6 \tab 15 \tab 3 \tab 18 \tab - \tab 3 \tab 5 \tab 13 \tab 24 \tab 11 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 6 \tab 2 \tab - \tab 1 \tab - \tab 5 \tab 4 \tab 1 \tab 2 \tab 18 \tab 15 \tab 5 \tab 12 \tab - \tab - \tab - \tab 2 \tab 24 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 8 \tab - \tab 29 \tab - \tab 5 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 5 \tab 4 \tab 2 \tab 1 \tab 4 \tab 3 \tab 8 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 4 \tab 10 \tab 21 \tab 1 \tab 9 \tab 6 \tab 2 \tab 8 \tab 4 \tab - \tab 1 \tab 6 \tab - \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab - \tab 9 \tab - \tab 2 \tab 5 \tab 1 \tab - \tab 6 \tab 1 \tab 14 \tab 4 \tab 9 \tab 5 \tab 4 \tab 16 \tab 1 \tab 2 \tab 1 \tab 11 \tab 2 \tab 12 \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 45 \tab 46 \tab - \tab - \tab 2 \tab 2 \tab - \tab 33 \tab 256 \tab 2 \tab 2 \tab 2 \tab 5 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 7 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1537 \cr
   1990 \tab - \tab 1 \tab 1 \tab - \tab 8 \tab 1 \tab 10 \tab 5 \tab 5 \tab 9 \tab 3 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab 8 \tab 11 \tab 3 \tab 3 \tab 5 \tab - \tab 14 \tab 5 \tab 3 \tab 6 \tab 13 \tab 1 \tab 8 \tab 45 \tab 4 \tab 4 \tab 1 \tab - \tab 5 \tab 5 \tab - \tab 3 \tab 3 \tab 2 \tab 13 \tab 9 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab 5 \tab 166 \tab 2 \tab - \tab - \tab - \tab 3 \tab 2 \tab 8 \tab - \tab 6 \tab 65 \tab 4 \tab 5 \tab 3 \tab 9 \tab 8 \tab 1 \tab 12 \tab - \tab 4 \tab 2 \tab 9 \tab 18 \tab 6 \tab 2 \tab 2 \tab - \tab 4 \tab - \tab 4 \tab 5 \tab 4 \tab 1 \tab 1 \tab - \tab 2 \tab 4 \tab 2 \tab 4 \tab 7 \tab 10 \tab 6 \tab 7 \tab 1 \tab - \tab 4 \tab 1 \tab 28 \tab 6 \tab 1 \tab 1 \tab - \tab - \tab 6 \tab 1 \tab 3 \tab 1 \tab 38 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 7 \tab 1 \tab 2 \tab 7 \tab 3 \tab 8 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 2 \tab 6 \tab 4 \tab 9 \tab 1 \tab 11 \tab 5 \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab 8 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 3 \tab - \tab 8 \tab 2 \tab 4 \tab 10 \tab 3 \tab 17 \tab - \tab - \tab 2 \tab 11 \tab - \tab 21 \tab 3 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 4 \tab 1 \tab 2 \tab - \tab 37 \tab 44 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 32 \tab 248 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 22 \tab 22 \tab 10 \tab 6 \tab 81 \tab 26 \tab 88 \tab 34 \tab 48 \tab 104 \tab 31 \tab 15 \tab 27 \tab 50 \tab 8 \tab 27 \tab 37 \tab 57 \tab 154 \tab 332 \tab 56 \tab 35 \tab 54 \tab 37 \tab 102 \tab 90 \tab 87 \tab 49 \tab 238 \tab 65 \tab 106 \tab 607 \tab 36 \tab 42 \tab 28 \tab 9 \tab 95 \tab 103 \tab 32 \tab 79 \tab 34 \tab 47 \tab 199 \tab 174 \tab 30 \tab 81 \tab 99 \tab 21 \tab 14 \tab 41 \tab 4056 \tab 25 \tab 5 \tab 23 \tab 23 \tab 27 \tab 49 \tab 153 \tab 3 \tab 77 \tab 1057 \tab 54 \tab 121 \tab 67 \tab 82 \tab 200 \tab 11 \tab 271 \tab 1 \tab 35 \tab 53 \tab 161 \tab 270 \tab 162 \tab 34 \tab 32 \tab 37 \tab 36 \tab 19 \tab 52 \tab 49 \tab 60 \tab 25 \tab 34 \tab 6 \tab 101 \tab 63 \tab 54 \tab 48 \tab 187 \tab 261 \tab 98 \tab 118 \tab 35 \tab 11 \tab 55 \tab 93 \tab 628 \tab 77 \tab 46 \tab 16 \tab 13 \tab 21 \tab 82 \tab 14 \tab 148 \tab 22 \tab 529 \tab 4 \tab 69 \tab 3 \tab 17 \tab 15 \tab 31 \tab 4 \tab 11 \tab 13 \tab 11 \tab 65 \tab 72 \tab 67 \tab 38 \tab 70 \tab 47 \tab 202 \tab 6 \tab 59 \tab 10 \tab 14 \tab 15 \tab 33 \tab 54 \tab 23 \tab 2 \tab 16 \tab 66 \tab 110 \tab 396 \tab 30 \tab 200 \tab 115 \tab 96 \tab 55 \tab 144 \tab 26 \tab 10 \tab 78 \tab 36 \tab 35 \tab 76 \tab 21 \tab 38 \tab 13 \tab 23 \tab 154 \tab 6 \tab 28 \tab 135 \tab 75 \tab 4 \tab 102 \tab 7 \tab 119 \tab 63 \tab 143 \tab 179 \tab 38 \tab 446 \tab 7 \tab 58 \tab 36 \tab 146 \tab 14 \tab 241 \tab 47 \tab 14 \tab 32 \tab 29 \tab 54 \tab 51 \tab 66 \tab 39 \tab 43 \tab 16 \tab 544 \tab 670 \tab 60 \tab 38 \tab 55 \tab 32 \tab 10 \tab 655 \tab 50385 \tab 36 \tab 16 \tab 10 \tab 44 \tab 2 \tab 9 \tab 11 \tab 20 \tab 86 \tab 17 \tab 11 \tab 16 \tab 11 \tab 14 \tab 14 \tab 6 \tab 2 \tab 5 \tab 2 \tab 29 \tab 6 \tab 9 \tab 9 \tab 5 \tab 11 \tab 7 \tab 20 \tab 6 \tab 22 \tab 7 \tab 12 \tab 10 \tab 15 \tab 6 \tab 64 \tab 15 \tab 10 \tab 14 \tab 15 \tab 5 \tab 9 \tab 14 \tab 3 \tab 2 \tab 3 \tab 17 \tab 11 \tab 25 \tab 4 \tab 2 \tab 17 \tab 4 \tab 7 \tab 14 \tab 3 \tab 20 \tab 10 \tab 11 \tab 4 \tab 2 \tab 3 \tab 3 \tab 14 \tab 5 \tab 6 \tab 5 \tab 2 \tab 1 \tab 1 \tab 5 \tab 14 \tab 2 \tab 7 \tab 13 \tab 28 \tab 5 \tab 15 \tab 4 \tab 11 \tab 5 \tab 10 \tab 15 \tab 9 \tab 1 \tab 4 \tab 4 \tab 3 \tab 4 \tab 4 \tab 8 \tab 2 \tab 4 \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
