% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_19.R
\name{padotged}
\alias{padotged}
\title{Fa. d.o.t. code - general educ develop.}
\source{
General Social Survey https://gss.norc.org
}
\description{
padotged
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 1 \tab 1.4 \tab 1.46 \tab 1.61 \tab 1.71 \tab 1.72 \tab 1.87 \tab 1.88 \tab 1.91 \tab 1.92 \tab 1.93 \tab 1.94 \tab 1.95 \tab 1.96 \tab 2.05 \tab 2.06 \tab 2.07 \tab 2.08 \tab 2.1 \tab 2.12 \tab 2.14 \tab 2.16 \tab 2.17 \tab 2.18 \tab 2.19 \tab 2.2 \tab 2.23 \tab 2.25 \tab 2.26 \tab 2.28 \tab 2.33 \tab 2.34 \tab 2.35 \tab 2.38 \tab 2.4 \tab 2.41 \tab 2.42 \tab 2.47 \tab 2.48 \tab 2.49 \tab 2.51 \tab 2.52 \tab 2.53 \tab 2.54 \tab 2.55 \tab 2.58 \tab 2.59 \tab 2.6 \tab 2.62 \tab 2.63 \tab 2.66 \tab 2.68 \tab 2.69 \tab 2.7 \tab 2.78 \tab 2.81 \tab 2.82 \tab 2.83 \tab 2.84 \tab 2.85 \tab 2.87 \tab 2.88 \tab 2.9 \tab 2.91 \tab 2.92 \tab 2.93 \tab 2.94 \tab 2.95 \tab 2.96 \tab 3 \tab 3.01 \tab 3.02 \tab 3.04 \tab 3.06 \tab 3.07 \tab 3.08 \tab 3.09 \tab 3.1 \tab 3.11 \tab 3.12 \tab 3.13 \tab 3.14 \tab 3.16 \tab 3.18 \tab 3.2 \tab 3.21 \tab 3.27 \tab 3.4 \tab 3.41 \tab 3.42 \tab 3.46 \tab 3.48 \tab 3.49 \tab 3.5 \tab 3.53 \tab 3.58 \tab 3.59 \tab 3.6 \tab 3.63 \tab 3.68 \tab 3.73 \tab 3.74 \tab 3.75 \tab 3.76 \tab 3.79 \tab 3.81 \tab 3.82 \tab 3.84 \tab 3.85 \tab 3.86 \tab 3.87 \tab 3.89 \tab 3.91 \tab 3.92 \tab 3.93 \tab 3.95 \tab 3.96 \tab 3.97 \tab 3.98 \tab 4 \tab 4.02 \tab 4.03 \tab 4.04 \tab 4.05 \tab 4.08 \tab 4.1 \tab 4.11 \tab 4.12 \tab 4.13 \tab 4.14 \tab 4.15 \tab 4.16 \tab 4.17 \tab 4.19 \tab 4.21 \tab 4.26 \tab 4.27 \tab 4.28 \tab 4.36 \tab 4.37 \tab 4.43 \tab 4.51 \tab 4.56 \tab 4.69 \tab 4.75 \tab 4.76 \tab 4.77 \tab 4.8 \tab 4.81 \tab 4.83 \tab 4.85 \tab 4.86 \tab 4.89 \tab 4.91 \tab 4.93 \tab 4.94 \tab 4.97 \tab 4.99 \tab 5 \tab 5.06 \tab 5.11 \tab 5.14 \tab 5.21 \tab 5.25 \tab 5.26 \tab 5.33 \tab 5.39 \tab 5.56 \tab 5.64 \tab 5.67 \tab 5.7 \tab 5.75 \tab 5.79 \tab 5.88 \tab 5.92 \tab 5.93 \tab 5.97 \tab 6 \tab iap \tab 1.85 \tab 2 \tab 2.04 \tab 2.43 \tab 2.57 \tab 2.61 \tab 2.67 \tab 2.75 \tab 2.86 \tab 2.89 \tab 3.03 \tab 3.05 \tab 3.26 \tab 3.3 \tab 3.57 \tab 3.8 \tab 3.83 \tab 3.94 \tab 3.99 \tab 4.01 \tab 4.33 \tab 4.39 \tab 4.42 \tab 4.7 \tab 4.98 \tab 5.04 \tab 5.07 \tab 5.13 \tab 5.28 \tab 5.47 \tab 5.48 \tab 5.84 \tab 1.21 \tab 1.98 \tab 2.29 \tab 2.77 \tab 2.99 \tab 3.17 \tab 4.25 \tab 5.1 \tab 5.38 \tab 5.55 \tab 5.59 \tab 1.86 \tab 2.24 \tab 2.37 \tab 2.39 \tab 2.46 \tab 2.74 \tab 3.38 \tab 3.69 \tab 4.09 \tab 4.32 \tab 4.96 \tab 5.29 \tab 2.02 \tab 3.77 \tab 4.22 \tab 4.46 \tab 4.95 \tab 1.82 \tab 4.63 \tab 5.95 \tab 1.75 \tab 2.45 \tab 2.64 \tab 2.22 \tab 2.98 \tab 3.19 \tab 4.47 \tab 4.62 \tab 4.88 \tab 5.23 \tab 1.76 \tab 5.83 \tab 1.84 \tab 2.21 \tab 4.06 \tab 5.91 \tab 3.47 \tab 3.9 \tab 1.26 \tab 1.7 \tab 3.33 \tab 3.65 \tab 4.24 \tab 4.4 \tab 3.32 \tab 4.82 \tab 5.19 \tab 2.36 \tab 4.68 \tab Total \cr
   1972 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab 7 \tab 1 \tab 3 \tab 3 \tab 7 \tab 1 \tab 3 \tab 15 \tab 2 \tab 3 \tab 1 \tab 1 \tab 11 \tab 9 \tab 2 \tab 4 \tab 5 \tab 1 \tab 1 \tab 1 \tab 13 \tab 4 \tab 6 \tab 2 \tab 6 \tab 4 \tab 4 \tab 46 \tab 3 \tab 3 \tab 2 \tab 1 \tab 7 \tab 2 \tab 3 \tab 1 \tab 2 \tab 2 \tab 3 \tab 2 \tab 2 \tab 7 \tab 40 \tab 4 \tab 3 \tab 6 \tab 4 \tab 12 \tab 2 \tab 3 \tab 1 \tab 29 \tab 5 \tab 11 \tab 18 \tab 10 \tab 16 \tab 2 \tab 2 \tab 5 \tab 11 \tab 14 \tab 4 \tab 7 \tab 8 \tab 5 \tab 2 \tab 9 \tab 4 \tab 1 \tab 8 \tab 8 \tab 22 \tab 2 \tab 5 \tab 3 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 14 \tab 2 \tab 9 \tab 1 \tab 2 \tab 8 \tab 7 \tab 16 \tab 10 \tab 3 \tab 314 \tab 8 \tab 4 \tab 35 \tab 4 \tab 15 \tab 5 \tab 1 \tab 24 \tab 11 \tab 5 \tab 6 \tab 13 \tab 1 \tab 7 \tab 6 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 5 \tab 2 \tab 13 \tab 7 \tab 1 \tab 2 \tab 9 \tab 1 \tab 9 \tab 2 \tab 2 \tab 77 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab 10 \tab 1 \tab 4 \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab 2 \tab 3 \tab 28 \tab 2 \tab 6 \tab 5 \tab 5 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 7 \tab 2 \tab 3 \tab 2 \tab 1 \tab 10 \tab 3 \tab 10 \tab 3 \tab 276 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 6 \tab 1 \tab 4 \tab 11 \tab - \tab 4 \tab 11 \tab - \tab - \tab 2 \tab 1 \tab 19 \tab 7 \tab 4 \tab 2 \tab 3 \tab 1 \tab 1 \tab 4 \tab 11 \tab 5 \tab 12 \tab 1 \tab - \tab 8 \tab - \tab 42 \tab 2 \tab 5 \tab 4 \tab 1 \tab 9 \tab - \tab 1 \tab - \tab 6 \tab 5 \tab 4 \tab 1 \tab - \tab 4 \tab 42 \tab 6 \tab 1 \tab 3 \tab 2 \tab 14 \tab 3 \tab 2 \tab 4 \tab 26 \tab 2 \tab 6 \tab 16 \tab 8 \tab 16 \tab - \tab 2 \tab 5 \tab 7 \tab 9 \tab - \tab 2 \tab 2 \tab 4 \tab 3 \tab 11 \tab 3 \tab 3 \tab 7 \tab 3 \tab 14 \tab 5 \tab 1 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 7 \tab 4 \tab 7 \tab 4 \tab 6 \tab 7 \tab 5 \tab 10 \tab 13 \tab - \tab 264 \tab 10 \tab 4 \tab 39 \tab 13 \tab 10 \tab 9 \tab 2 \tab 23 \tab 11 \tab 9 \tab 48 \tab 5 \tab 1 \tab 6 \tab 5 \tab 5 \tab 4 \tab 6 \tab - \tab 4 \tab 5 \tab 6 \tab 11 \tab 7 \tab 2 \tab 2 \tab 7 \tab 2 \tab 9 \tab 2 \tab - \tab 82 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 15 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 8 \tab 20 \tab 5 \tab 4 \tab 4 \tab 5 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 5 \tab 1 \tab 5 \tab 5 \tab - \tab 4 \tab 3 \tab - \tab 8 \tab 2 \tab 12 \tab 1 \tab 191 \tab 1 \tab 4 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab - \tab 2 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab 2 \tab 6 \tab - \tab 1 \tab 10 \tab 1 \tab 1 \tab - \tab 3 \tab 15 \tab 5 \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 2 \tab 3 \tab 8 \tab 8 \tab 1 \tab 2 \tab 3 \tab 1 \tab 43 \tab 1 \tab 5 \tab 4 \tab 3 \tab 15 \tab - \tab 4 \tab - \tab 1 \tab 7 \tab 1 \tab 3 \tab 1 \tab 6 \tab 30 \tab 9 \tab - \tab 9 \tab 4 \tab 23 \tab 2 \tab 1 \tab 1 \tab 18 \tab 6 \tab 8 \tab 13 \tab 11 \tab 10 \tab 1 \tab 6 \tab 3 \tab 6 \tab 9 \tab 2 \tab 6 \tab 6 \tab 1 \tab 9 \tab 9 \tab 2 \tab 4 \tab 6 \tab 4 \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 5 \tab 1 \tab - \tab 5 \tab - \tab 11 \tab 5 \tab 10 \tab 5 \tab 4 \tab 8 \tab 2 \tab 21 \tab 29 \tab - \tab 275 \tab 15 \tab 5 \tab 44 \tab 9 \tab 15 \tab 2 \tab 1 \tab 14 \tab 5 \tab 12 \tab 6 \tab 9 \tab - \tab 5 \tab 6 \tab 1 \tab 2 \tab 7 \tab 1 \tab 3 \tab 7 \tab 2 \tab 9 \tab 9 \tab - \tab 2 \tab 8 \tab - \tab 2 \tab 2 \tab 1 \tab 80 \tab - \tab 3 \tab - \tab 3 \tab 3 \tab 13 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 2 \tab 7 \tab 27 \tab 4 \tab 5 \tab 8 \tab 2 \tab - \tab - \tab - \tab 2 \tab 5 \tab - \tab 3 \tab 2 \tab 2 \tab 3 \tab 2 \tab 8 \tab - \tab - \tab 5 \tab 5 \tab 7 \tab 2 \tab 215 \tab 2 \tab 6 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 4 \tab - \tab - \tab 3 \tab 5 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1 \tab - \tab 3 \tab - \tab - \tab 3 \tab 5 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 5 \tab - \tab 2 \tab 14 \tab 1 \tab - \tab 2 \tab 4 \tab 22 \tab 4 \tab 5 \tab 3 \tab 2 \tab - \tab - \tab - \tab 2 \tab 14 \tab 4 \tab 1 \tab 1 \tab - \tab 2 \tab 27 \tab 3 \tab 2 \tab 4 \tab 1 \tab 15 \tab - \tab 3 \tab - \tab 1 \tab 9 \tab 1 \tab 1 \tab 1 \tab 3 \tab 35 \tab 6 \tab - \tab 10 \tab 8 \tab 36 \tab 1 \tab - \tab 4 \tab 32 \tab 4 \tab 5 \tab 12 \tab 10 \tab 14 \tab 2 \tab - \tab 8 \tab 7 \tab 15 \tab 1 \tab 10 \tab 3 \tab 1 \tab 1 \tab 10 \tab 3 \tab 1 \tab 7 \tab 6 \tab 29 \tab 3 \tab 2 \tab - \tab 1 \tab 4 \tab 2 \tab - \tab 2 \tab 1 \tab 4 \tab 4 \tab 17 \tab 3 \tab 1 \tab 5 \tab 5 \tab 23 \tab 17 \tab - \tab 266 \tab 7 \tab - \tab 44 \tab 4 \tab 14 \tab 1 \tab 1 \tab 40 \tab 13 \tab 8 \tab 15 \tab 12 \tab - \tab 7 \tab 9 \tab - \tab 3 \tab 7 \tab - \tab 3 \tab 1 \tab 1 \tab 12 \tab 2 \tab 1 \tab - \tab 9 \tab 1 \tab 5 \tab 3 \tab 2 \tab 71 \tab 2 \tab - \tab 3 \tab 3 \tab - \tab 9 \tab 2 \tab 4 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 4 \tab 6 \tab 26 \tab 3 \tab 3 \tab 6 \tab 7 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab 8 \tab 6 \tab 1 \tab 3 \tab 1 \tab - \tab 8 \tab 2 \tab 8 \tab 2 \tab 190 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 3 \tab 2 \tab 2 \tab 5 \tab - \tab 4 \tab 5 \tab - \tab - \tab 20 \tab 2 \tab 1 \tab 1 \tab 3 \tab 19 \tab 7 \tab 2 \tab 2 \tab 4 \tab - \tab - \tab - \tab 1 \tab 15 \tab 16 \tab 1 \tab 2 \tab 2 \tab 4 \tab 22 \tab - \tab 1 \tab 7 \tab 2 \tab 11 \tab 1 \tab 3 \tab 1 \tab 11 \tab 5 \tab 4 \tab 5 \tab 3 \tab 2 \tab 40 \tab 8 \tab 1 \tab 9 \tab 2 \tab 24 \tab 3 \tab 2 \tab 7 \tab 28 \tab 2 \tab 7 \tab 11 \tab 5 \tab 12 \tab 1 \tab 2 \tab 3 \tab 7 \tab 3 \tab 3 \tab 3 \tab 2 \tab 2 \tab 1 \tab 8 \tab 1 \tab 4 \tab 8 \tab 9 \tab 28 \tab 3 \tab 2 \tab 4 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 12 \tab 2 \tab 8 \tab 2 \tab 1 \tab 15 \tab 6 \tab 19 \tab 13 \tab - \tab 288 \tab 16 \tab 1 \tab 35 \tab 6 \tab 3 \tab 4 \tab 4 \tab 18 \tab 13 \tab 7 \tab 13 \tab 8 \tab - \tab 5 \tab 4 \tab 1 \tab 3 \tab 5 \tab - \tab 3 \tab 3 \tab 1 \tab 22 \tab 2 \tab 1 \tab - \tab 11 \tab 5 \tab 6 \tab 10 \tab 2 \tab 60 \tab - \tab 1 \tab - \tab 3 \tab - \tab 8 \tab 3 \tab 2 \tab 1 \tab 2 \tab 4 \tab 4 \tab 1 \tab - \tab 1 \tab 33 \tab 3 \tab 8 \tab 8 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 5 \tab - \tab 2 \tab - \tab - \tab 13 \tab 2 \tab 9 \tab 4 \tab 219 \tab 1 \tab 3 \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 5 \tab - \tab 2 \tab 16 \tab 1 \tab - \tab 2 \tab 3 \tab 20 \tab 9 \tab 4 \tab 1 \tab 5 \tab 3 \tab 1 \tab 2 \tab 2 \tab 10 \tab 5 \tab - \tab 3 \tab 2 \tab 3 \tab 50 \tab 1 \tab 4 \tab 6 \tab 4 \tab 19 \tab - \tab 3 \tab - \tab 3 \tab 4 \tab 6 \tab 3 \tab 2 \tab 2 \tab 59 \tab 4 \tab 2 \tab 6 \tab 2 \tab 31 \tab 2 \tab 2 \tab 6 \tab 26 \tab 7 \tab 7 \tab 8 \tab 11 \tab 21 \tab 2 \tab 1 \tab 2 \tab 7 \tab 4 \tab 5 \tab 6 \tab 8 \tab 1 \tab 2 \tab 11 \tab 2 \tab 4 \tab 6 \tab 5 \tab 21 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 9 \tab 2 \tab 9 \tab - \tab - \tab 9 \tab 9 \tab 18 \tab 17 \tab 2 \tab 262 \tab 22 \tab 4 \tab 30 \tab 7 \tab 8 \tab 5 \tab 1 \tab 18 \tab 17 \tab 9 \tab 9 \tab 7 \tab - \tab 6 \tab 4 \tab 4 \tab 3 \tab 3 \tab - \tab 2 \tab 2 \tab 1 \tab 11 \tab 2 \tab - \tab - \tab 10 \tab 1 \tab 5 \tab 4 \tab 1 \tab 62 \tab - \tab - \tab 3 \tab 3 \tab - \tab 9 \tab - \tab 2 \tab 1 \tab 6 \tab 1 \tab 2 \tab 1 \tab 6 \tab 3 \tab 37 \tab 5 \tab 3 \tab 4 \tab 8 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 8 \tab - \tab - \tab 14 \tab 3 \tab 9 \tab 1 \tab 236 \tab - \tab 1 \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 2 \tab - \tab - \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 4 \tab 3 \tab 5 \tab 11 \tab - \tab 2 \tab 14 \tab 1 \tab - \tab 1 \tab 3 \tab 14 \tab 2 \tab 5 \tab 3 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 9 \tab 5 \tab 2 \tab 2 \tab - \tab 3 \tab 33 \tab 3 \tab 1 \tab 6 \tab 3 \tab 18 \tab 1 \tab - \tab 2 \tab 9 \tab 3 \tab 1 \tab 3 \tab 2 \tab 5 \tab 56 \tab 2 \tab 2 \tab 9 \tab 2 \tab 31 \tab 1 \tab 1 \tab 5 \tab 27 \tab 6 \tab 6 \tab 12 \tab 18 \tab 16 \tab 3 \tab 2 \tab 2 \tab 8 \tab 10 \tab 1 \tab 6 \tab 10 \tab 2 \tab 2 \tab 10 \tab 1 \tab 3 \tab 10 \tab 3 \tab 35 \tab 1 \tab 5 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 7 \tab - \tab 5 \tab 5 \tab 1 \tab 7 \tab 2 \tab 12 \tab 14 \tab 1 \tab 241 \tab 21 \tab 1 \tab 54 \tab 8 \tab 9 \tab 2 \tab 2 \tab 12 \tab 30 \tab 9 \tab 11 \tab 8 \tab 2 \tab 6 \tab 2 \tab 1 \tab 1 \tab 6 \tab - \tab 1 \tab 4 \tab 6 \tab 8 \tab 1 \tab 1 \tab 1 \tab 10 \tab 5 \tab 5 \tab 5 \tab - \tab 70 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 12 \tab - \tab 2 \tab 2 \tab 6 \tab - \tab 2 \tab 4 \tab 3 \tab 2 \tab 35 \tab 5 \tab 10 \tab 7 \tab 5 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab 11 \tab 1 \tab 5 \tab 1 \tab - \tab 15 \tab - \tab 6 \tab - \tab 244 \tab - \tab 2 \tab - \tab - \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 6 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab 3 \tab 1 \tab 4 \tab 7 \tab - \tab 2 \tab 10 \tab 2 \tab - \tab - \tab 3 \tab 14 \tab 4 \tab 2 \tab 6 \tab 1 \tab - \tab - \tab - \tab 4 \tab 7 \tab 6 \tab 3 \tab 4 \tab 1 \tab 1 \tab 20 \tab 1 \tab 1 \tab 8 \tab - \tab 11 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 50 \tab 10 \tab 2 \tab 7 \tab 2 \tab 26 \tab 1 \tab 1 \tab 3 \tab 22 \tab 3 \tab 3 \tab 10 \tab 7 \tab 13 \tab 2 \tab - \tab 5 \tab 8 \tab 13 \tab 1 \tab 6 \tab 2 \tab 2 \tab 1 \tab 11 \tab 1 \tab 3 \tab 8 \tab 3 \tab 26 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 4 \tab - \tab 4 \tab 12 \tab 9 \tab 6 \tab - \tab 4 \tab 2 \tab 13 \tab 11 \tab 2 \tab 234 \tab 13 \tab 6 \tab 40 \tab 5 \tab 27 \tab 2 \tab 4 \tab 22 \tab 12 \tab 6 \tab 9 \tab 13 \tab 2 \tab 6 \tab 5 \tab 2 \tab 4 \tab 4 \tab - \tab 1 \tab 9 \tab 2 \tab 10 \tab 1 \tab 1 \tab 4 \tab 10 \tab 8 \tab 3 \tab 5 \tab 1 \tab 38 \tab 5 \tab - \tab 2 \tab 1 \tab 2 \tab 12 \tab - \tab 3 \tab 4 \tab 2 \tab 1 \tab 6 \tab 5 \tab 1 \tab 4 \tab 24 \tab 2 \tab 5 \tab 13 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 6 \tab - \tab 7 \tab 1 \tab - \tab 17 \tab - \tab 5 \tab 9 \tab 260 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab 4 \tab 2 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 6 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 2 \tab - \tab 5 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 11 \tab - \tab - \tab 18 \tab - \tab 3 \tab 3 \tab 4 \tab 43 \tab 10 \tab 4 \tab 2 \tab 4 \tab - \tab - \tab 2 \tab 5 \tab 13 \tab 13 \tab 4 \tab 1 \tab 1 \tab 2 \tab 52 \tab 6 \tab 8 \tab 4 \tab 2 \tab 13 \tab 2 \tab 4 \tab 2 \tab 11 \tab 12 \tab 6 \tab - \tab 1 \tab 7 \tab 59 \tab 4 \tab 2 \tab 9 \tab 3 \tab 22 \tab 1 \tab 3 \tab 3 \tab 48 \tab 7 \tab 6 \tab 12 \tab 23 \tab 16 \tab 2 \tab 2 \tab 7 \tab 16 \tab 13 \tab - \tab 7 \tab 6 \tab 4 \tab 1 \tab 6 \tab 3 \tab 3 \tab 11 \tab 6 \tab 48 \tab 5 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 11 \tab 10 \tab 9 \tab 1 \tab 1 \tab - \tab 4 \tab 17 \tab 7 \tab 1 \tab 273 \tab 14 \tab 19 \tab 48 \tab 5 \tab 28 \tab 7 \tab 2 \tab 18 \tab 16 \tab 12 \tab 8 \tab 11 \tab 6 \tab 7 \tab 6 \tab 1 \tab 8 \tab 4 \tab - \tab 1 \tab 2 \tab 6 \tab 3 \tab 6 \tab 2 \tab 1 \tab 11 \tab 11 \tab 3 \tab 4 \tab 2 \tab 24 \tab - \tab - \tab 1 \tab 3 \tab - \tab 5 \tab 1 \tab 4 \tab - \tab 5 \tab - \tab 2 \tab 2 \tab 2 \tab 4 \tab 23 \tab 6 \tab 10 \tab 3 \tab 6 \tab - \tab 5 \tab 7 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 4 \tab - \tab 4 \tab 1 \tab - \tab 10 \tab 3 \tab 13 \tab - \tab 378 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 12 \tab 2 \tab 3 \tab 2 \tab 2 \tab 7 \tab - \tab - \tab 8 \tab 1 \tab 2 \tab 2 \tab - \tab 9 \tab 7 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 8 \tab 9 \tab 5 \tab 1 \tab 7 \tab 1 \tab - \tab 29 \tab 4 \tab - \tab 6 \tab - \tab 15 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 6 \tab 4 \tab 6 \tab 7 \tab 36 \tab 9 \tab 1 \tab 6 \tab 7 \tab 28 \tab 1 \tab 3 \tab 7 \tab 36 \tab 9 \tab 7 \tab 17 \tab 19 \tab 18 \tab 3 \tab 1 \tab 5 \tab 15 \tab 12 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab 10 \tab 2 \tab 6 \tab 5 \tab 5 \tab 41 \tab 3 \tab 3 \tab - \tab 4 \tab 6 \tab - \tab - \tab 3 \tab - \tab 5 \tab 13 \tab 8 \tab 5 \tab - \tab 4 \tab 6 \tab 24 \tab 13 \tab 6 \tab 194 \tab 14 \tab 11 \tab 49 \tab 5 \tab 15 \tab 6 \tab 5 \tab 22 \tab 18 \tab 3 \tab 7 \tab 17 \tab 3 \tab 1 \tab 3 \tab 3 \tab 5 \tab 7 \tab 1 \tab 4 \tab 3 \tab 5 \tab 19 \tab 1 \tab 1 \tab 6 \tab 8 \tab 4 \tab 7 \tab 6 \tab 1 \tab 36 \tab 4 \tab 3 \tab - \tab 5 \tab 1 \tab 15 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 4 \tab 7 \tab 2 \tab 2 \tab 43 \tab 6 \tab 7 \tab 9 \tab 12 \tab 1 \tab 4 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 3 \tab 10 \tab 8 \tab - \tab 5 \tab 3 \tab - \tab 18 \tab 6 \tab 12 \tab 5 \tab 238 \tab - \tab 4 \tab - \tab 2 \tab 3 \tab - \tab - \tab 3 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 5 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 7 \tab 5 \tab - \tab - \tab 6 \tab - \tab - \tab 5 \tab - \tab 24 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 58 \tab 6 \tab - \tab 1 \tab 3 \tab - \tab - \tab 29 \tab 2 \tab 2 \tab 6 \tab - \tab 31 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 31 \tab 9 \tab 3 \tab 3 \tab 5 \tab 1 \tab 16 \tab - \tab - \tab - \tab 34 \tab 11 \tab 4 \tab 14 \tab 14 \tab 13 \tab 1 \tab 2 \tab 4 \tab 10 \tab 8 \tab 3 \tab - \tab 1 \tab 2 \tab 3 \tab 8 \tab - \tab 2 \tab 6 \tab 4 \tab 32 \tab 8 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 10 \tab 19 \tab 1 \tab 4 \tab - \tab 3 \tab 4 \tab 8 \tab 4 \tab 3 \tab 233 \tab 11 \tab 4 \tab 21 \tab 2 \tab 14 \tab 2 \tab 2 \tab 12 \tab 16 \tab 6 \tab 14 \tab 8 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 4 \tab - \tab 28 \tab - \tab - \tab - \tab 12 \tab 5 \tab - \tab 1 \tab 3 \tab 101 \tab 2 \tab - \tab - \tab 3 \tab 4 \tab 5 \tab 1 \tab 2 \tab - \tab 6 \tab - \tab - \tab 6 \tab 1 \tab 5 \tab 52 \tab 4 \tab 6 \tab 6 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab 3 \tab 8 \tab 2 \tab 6 \tab 2 \tab - \tab 8 \tab 1 \tab 5 \tab 3 \tab 249 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab 4 \tab 1 \tab 4 \tab 1 \tab - \tab 2 \tab 6 \tab 4 \tab - \tab 1 \tab 1 \tab 26 \tab 5 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 4 \tab 2 \tab 4 \tab 4 \tab - \tab 6 \tab 2 \tab - \tab 30 \tab 7 \tab 4 \tab 6 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 11 \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 26 \tab 4 \tab 2 \tab 6 \tab 3 \tab 16 \tab - \tab - \tab 1 \tab 38 \tab 5 \tab 9 \tab 13 \tab 11 \tab 8 \tab 2 \tab 2 \tab 6 \tab 17 \tab 5 \tab 2 \tab 9 \tab 9 \tab 1 \tab - \tab 12 \tab - \tab 1 \tab 7 \tab 1 \tab 29 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 8 \tab 13 \tab 3 \tab - \tab 2 \tab 3 \tab 22 \tab 12 \tab 1 \tab 226 \tab 16 \tab 8 \tab 32 \tab 4 \tab 16 \tab 3 \tab 3 \tab 20 \tab 24 \tab 9 \tab 19 \tab 11 \tab 5 \tab 7 \tab 11 \tab 1 \tab 9 \tab 6 \tab 1 \tab 5 \tab 6 \tab 6 \tab 14 \tab 6 \tab 5 \tab 4 \tab 14 \tab 6 \tab 5 \tab 8 \tab 5 \tab 59 \tab 1 \tab 5 \tab 1 \tab 2 \tab 4 \tab 9 \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 2 \tab 3 \tab 28 \tab 6 \tab 6 \tab 8 \tab 10 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 2 \tab 3 \tab 10 \tab 6 \tab 5 \tab 10 \tab - \tab - \tab 17 \tab 3 \tab 14 \tab 3 \tab 219 \tab 1 \tab 5 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 3 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 7 \tab 2 \tab 1 \tab - \tab 8 \tab 2 \tab 1 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 3 \tab 10 \tab 6 \tab 3 \tab 7 \tab 2 \tab - \tab 1 \tab 2 \tab 5 \tab 3 \tab 4 \tab - \tab 1 \tab 3 \tab 3 \tab 28 \tab - \tab 9 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 6 \tab 1 \tab 3 \tab 2 \tab 2 \tab 3 \tab 23 \tab 6 \tab - \tab 3 \tab 1 \tab 19 \tab - \tab - \tab 2 \tab 33 \tab 8 \tab 3 \tab 22 \tab 14 \tab 18 \tab 3 \tab 2 \tab 2 \tab 3 \tab 5 \tab 1 \tab 7 \tab 3 \tab - \tab 1 \tab 12 \tab - \tab 1 \tab 5 \tab 6 \tab 55 \tab 4 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 6 \tab 8 \tab 7 \tab 2 \tab 1 \tab - \tab 1 \tab 7 \tab 13 \tab 2 \tab 219 \tab 11 \tab 4 \tab 33 \tab 6 \tab 27 \tab 3 \tab - \tab 34 \tab 20 \tab 3 \tab 3 \tab 12 \tab 8 \tab 6 \tab 7 \tab 5 \tab 3 \tab 7 \tab - \tab 1 \tab 6 \tab 2 \tab 14 \tab 1 \tab 4 \tab 7 \tab 6 \tab 5 \tab 4 \tab 5 \tab 3 \tab 54 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 5 \tab 3 \tab - \tab 2 \tab 3 \tab 3 \tab 3 \tab - \tab 4 \tab 2 \tab 34 \tab 4 \tab 3 \tab 11 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 4 \tab 8 \tab 2 \tab 5 \tab 1 \tab 2 \tab 12 \tab 3 \tab 8 \tab 7 \tab 258 \tab 3 \tab 5 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 7 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 6 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab 1 \tab 2 \tab 6 \tab 3 \tab - \tab - \tab 3 \tab - \tab 3 \tab 3 \tab 4 \tab 1 \tab 6 \tab 6 \tab 2 \tab 4 \tab 6 \tab - \tab - \tab 2 \tab 2 \tab 42 \tab 5 \tab 2 \tab 2 \tab 6 \tab - \tab - \tab 1 \tab 7 \tab 9 \tab 11 \tab 1 \tab 2 \tab 2 \tab 3 \tab 46 \tab 4 \tab 2 \tab 10 \tab 2 \tab 9 \tab - \tab 3 \tab 1 \tab 9 \tab 2 \tab 13 \tab 2 \tab 1 \tab 12 \tab 41 \tab 8 \tab 4 \tab 4 \tab - \tab 39 \tab 4 \tab 1 \tab 8 \tab 43 \tab - \tab 4 \tab 15 \tab 13 \tab 15 \tab 3 \tab 2 \tab 3 \tab 11 \tab 5 \tab 6 \tab 3 \tab 3 \tab 1 \tab - \tab 9 \tab 1 \tab 3 \tab 10 \tab 5 \tab 35 \tab 9 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 13 \tab 14 \tab 3 \tab 6 \tab 2 \tab 1 \tab 3 \tab 13 \tab 15 \tab - \tab 272 \tab 8 \tab 7 \tab 45 \tab 4 \tab 18 \tab 2 \tab 1 \tab 14 \tab 19 \tab 5 \tab 7 \tab 10 \tab 7 \tab 2 \tab 5 \tab 2 \tab 6 \tab 7 \tab 1 \tab 3 \tab 7 \tab 5 \tab 11 \tab 2 \tab 1 \tab 4 \tab 10 \tab 7 \tab 6 \tab 6 \tab 4 \tab 35 \tab 2 \tab 1 \tab - \tab 2 \tab 8 \tab 19 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab 4 \tab 34 \tab 3 \tab 12 \tab 5 \tab 5 \tab - \tab 9 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 8 \tab 13 \tab 1 \tab 4 \tab 3 \tab 1 \tab 20 \tab 2 \tab 11 \tab 4 \tab 341 \tab 2 \tab 5 \tab - \tab 1 \tab 4 \tab 3 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 6 \tab 1 \tab - \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 6 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 4 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab - \tab 5 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 5 \tab 4 \tab - \tab - \tab 5 \tab - \tab - \tab 2 \tab 2 \tab 28 \tab 11 \tab - \tab 4 \tab 3 \tab - \tab - \tab - \tab 7 \tab 8 \tab 2 \tab - \tab 5 \tab - \tab - \tab 27 \tab 3 \tab 7 \tab 8 \tab 3 \tab 7 \tab - \tab 1 \tab - \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 3 \tab 15 \tab 3 \tab 1 \tab - \tab 3 \tab 20 \tab 1 \tab - \tab 2 \tab 36 \tab 5 \tab 8 \tab 20 \tab 16 \tab 12 \tab 3 \tab 1 \tab 3 \tab 6 \tab 7 \tab 1 \tab 3 \tab 6 \tab 5 \tab 1 \tab 9 \tab 3 \tab 1 \tab 9 \tab 3 \tab 44 \tab 4 \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 11 \tab 2 \tab 12 \tab 3 \tab - \tab 3 \tab - \tab 19 \tab 12 \tab 5 \tab 205 \tab 6 \tab 5 \tab 29 \tab 6 \tab 19 \tab 4 \tab 3 \tab 24 \tab 19 \tab 4 \tab 11 \tab 9 \tab 3 \tab 2 \tab 2 \tab 5 \tab 6 \tab 5 \tab - \tab 5 \tab 5 \tab 5 \tab 7 \tab 1 \tab - \tab 2 \tab 8 \tab 7 \tab 5 \tab 2 \tab 4 \tab 56 \tab 2 \tab 2 \tab 8 \tab 3 \tab 1 \tab 13 \tab 1 \tab 1 \tab - \tab 6 \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab 40 \tab 6 \tab 7 \tab 5 \tab 3 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 2 \tab 8 \tab 7 \tab 5 \tab 12 \tab 5 \tab 2 \tab 11 \tab 3 \tab 10 \tab 7 \tab 242 \tab 5 \tab 4 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 3 \tab 8 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1481 \cr
   1989 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 2 \tab 15 \tab - \tab 1 \tab 1 \tab 4 \tab 30 \tab 7 \tab - \tab 4 \tab - \tab - \tab 3 \tab 2 \tab 5 \tab 11 \tab 6 \tab 1 \tab 3 \tab 2 \tab - \tab 23 \tab 4 \tab 1 \tab 6 \tab 1 \tab 6 \tab - \tab 1 \tab 1 \tab 8 \tab - \tab 9 \tab 1 \tab 1 \tab 2 \tab 30 \tab 2 \tab 1 \tab 5 \tab 4 \tab 18 \tab 3 \tab 1 \tab 3 \tab 45 \tab 4 \tab 8 \tab 12 \tab 18 \tab 9 \tab 1 \tab - \tab 6 \tab 8 \tab 7 \tab - \tab 8 \tab 3 \tab 3 \tab - \tab 11 \tab - \tab 6 \tab 6 \tab 2 \tab 29 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab 4 \tab 9 \tab 3 \tab 2 \tab 4 \tab 4 \tab 26 \tab 13 \tab 1 \tab 204 \tab 15 \tab 6 \tab 24 \tab 7 \tab 8 \tab 5 \tab 2 \tab 30 \tab 21 \tab 6 \tab 7 \tab 8 \tab 1 \tab 8 \tab 4 \tab 3 \tab 6 \tab - \tab - \tab - \tab 3 \tab 5 \tab 15 \tab 1 \tab 2 \tab 5 \tab 5 \tab 6 \tab 4 \tab 7 \tab 3 \tab 67 \tab 2 \tab 3 \tab 1 \tab 2 \tab 2 \tab 10 \tab 6 \tab 3 \tab 2 \tab 2 \tab 3 \tab 3 \tab 4 \tab 5 \tab 6 \tab 48 \tab 9 \tab 13 \tab 8 \tab 7 \tab 1 \tab 5 \tab 2 \tab - \tab 3 \tab 1 \tab 2 \tab 4 \tab 7 \tab 10 \tab 3 \tab 9 \tab 2 \tab 1 \tab 12 \tab - \tab 12 \tab 5 \tab 256 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 8 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1537 \cr
   1990 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 3 \tab - \tab 3 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab 7 \tab 3 \tab - \tab - \tab 1 \tab 29 \tab 8 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 11 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab 21 \tab 5 \tab 3 \tab 3 \tab 1 \tab 5 \tab - \tab 2 \tab - \tab 5 \tab 1 \tab 3 \tab 2 \tab 3 \tab 2 \tab 22 \tab 7 \tab - \tab 8 \tab 1 \tab 27 \tab 2 \tab - \tab 2 \tab 37 \tab 2 \tab 6 \tab 5 \tab 10 \tab 17 \tab 2 \tab 4 \tab 3 \tab 3 \tab 6 \tab 1 \tab 2 \tab 5 \tab 3 \tab - \tab 6 \tab 1 \tab 2 \tab 4 \tab 1 \tab 38 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 10 \tab 1 \tab 11 \tab 2 \tab 1 \tab 2 \tab 1 \tab 19 \tab 10 \tab - \tab 167 \tab 14 \tab 3 \tab 28 \tab 3 \tab 10 \tab 3 \tab 2 \tab 16 \tab 17 \tab 5 \tab 8 \tab 5 \tab 6 \tab 6 \tab 8 \tab 7 \tab 5 \tab 3 \tab - \tab 2 \tab 8 \tab 1 \tab 9 \tab - \tab 3 \tab 5 \tab 9 \tab 8 \tab 4 \tab 6 \tab 1 \tab 64 \tab - \tab - \tab 3 \tab 2 \tab 4 \tab 9 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 4 \tab 3 \tab 3 \tab 1 \tab 41 \tab 8 \tab 8 \tab 13 \tab 5 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 3 \tab 5 \tab 10 \tab 9 \tab 1 \tab 6 \tab 1 \tab 2 \tab 7 \tab 2 \tab 8 \tab 4 \tab 248 \tab - \tab 6 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 11 \tab 5 \tab 34 \tab 14 \tab 5 \tab 18 \tab 21 \tab 6 \tab 19 \tab 32 \tab 9 \tab 11 \tab 13 \tab 20 \tab 38 \tab 40 \tab 45 \tab 18 \tab 62 \tab 98 \tab 4 \tab 26 \tab 186 \tab 18 \tab 12 \tab 26 \tab 38 \tab 375 \tab 109 \tab 39 \tab 46 \tab 49 \tab 9 \tab 14 \tab 22 \tab 137 \tab 146 \tab 112 \tab 20 \tab 49 \tab 31 \tab 26 \tab 568 \tab 49 \tab 58 \tab 93 \tab 25 \tab 198 \tab 10 \tab 35 \tab 9 \tab 86 \tab 60 \tab 72 \tab 30 \tab 29 \tab 100 \tab 613 \tab 95 \tab 25 \tab 105 \tab 49 \tab 402 \tab 27 \tab 20 \tab 59 \tab 558 \tab 86 \tab 108 \tab 230 \tab 218 \tab 244 \tab 33 \tab 31 \tab 72 \tab 150 \tab 145 \tab 34 \tab 87 \tab 79 \tab 37 \tab 27 \tab 162 \tab 27 \tab 48 \tab 123 \tab 74 \tab 529 \tab 59 \tab 30 \tab 21 \tab 22 \tab 38 \tab 13 \tab 3 \tab 31 \tab 8 \tab 142 \tab 110 \tab 147 \tab 55 \tab 22 \tab 82 \tab 64 \tab 287 \tab 223 \tab 27 \tab 4137 \tab 221 \tab 92 \tab 630 \tab 98 \tab 256 \tab 65 \tab 36 \tab 361 \tab 282 \tab 118 \tab 201 \tab 166 \tab 45 \tab 90 \tab 89 \tab 42 \tab 71 \tab 79 \tab 6 \tab 39 \tab 80 \tab 56 \tab 216 \tab 49 \tab 25 \tab 45 \tab 157 \tab 82 \tab 82 \tab 78 \tab 35 \tab 1036 \tab 27 \tab 23 \tab 27 \tab 41 \tab 38 \tab 178 \tab 22 \tab 40 \tab 20 \tab 49 \tab 28 \tab 44 \tab 41 \tab 38 \tab 65 \tab 573 \tab 81 \tab 116 \tab 123 \tab 92 \tab 10 \tab 28 \tab 40 \tab 14 \tab 23 \tab 9 \tab 39 \tab 34 \tab 92 \tab 117 \tab 27 \tab 101 \tab 26 \tab 9 \tab 205 \tab 40 \tab 159 \tab 60 \tab 50385 \tab 16 \tab 55 \tab 1 \tab 10 \tab 37 \tab 9 \tab 15 \tab 15 \tab 19 \tab 36 \tab 15 \tab 3 \tab 8 \tab 9 \tab 29 \tab 2 \tab 15 \tab 67 \tab 10 \tab 9 \tab 17 \tab 5 \tab 15 \tab 16 \tab 4 \tab 20 \tab 23 \tab 12 \tab 10 \tab 36 \tab 14 \tab 6 \tab 4 \tab 8 \tab 6 \tab 3 \tab 19 \tab 14 \tab 5 \tab 19 \tab 3 \tab 7 \tab 11 \tab 5 \tab 7 \tab 9 \tab 1 \tab 2 \tab 3 \tab 1 \tab 9 \tab 2 \tab 4 \tab 17 \tab 14 \tab 2 \tab 2 \tab 15 \tab 28 \tab 7 \tab 9 \tab 9 \tab 13 \tab 4 \tab 5 \tab 1 \tab 7 \tab 9 \tab 2 \tab 4 \tab 10 \tab 3 \tab 15 \tab 5 \tab 1 \tab 3 \tab 4 \tab 8 \tab 1 \tab 3 \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
