% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_19.R
\name{padotpeo}
\alias{padotpeo}
\title{padotpeo}
\source{
General Social Survey https://gss.norc.org
}
\description{
Fa. d.o.t. code - relation to people
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 0 \tab 0.06 \tab 0.42 \tab 0.54 \tab 0.62 \tab 1.43 \tab 1.89 \tab 1.97 \tab 2 \tab 2.02 \tab 2.03 \tab 3.06 \tab 3.33 \tab 3.55 \tab 3.68 \tab 3.75 \tab 3.77 \tab 3.85 \tab 3.94 \tab 3.95 \tab 3.96 \tab 3.97 \tab 4 \tab 4.02 \tab 4.04 \tab 4.15 \tab 4.28 \tab 4.36 \tab 4.39 \tab 4.54 \tab 4.72 \tab 4.84 \tab 4.86 \tab 4.94 \tab 5.01 \tab 5.02 \tab 5.07 \tab 5.09 \tab 5.13 \tab 5.14 \tab 5.17 \tab 5.26 \tab 5.33 \tab 5.36 \tab 5.38 \tab 5.42 \tab 5.43 \tab 5.45 \tab 5.48 \tab 5.5 \tab 5.57 \tab 5.58 \tab 5.59 \tab 5.62 \tab 5.65 \tab 5.67 \tab 5.69 \tab 5.78 \tab 5.84 \tab 5.86 \tab 5.88 \tab 5.9 \tab 5.95 \tab 5.99 \tab 6 \tab 6.02 \tab 6.03 \tab 6.04 \tab 6.05 \tab 6.06 \tab 6.09 \tab 6.15 \tab 6.17 \tab 6.28 \tab 6.45 \tab 6.46 \tab 6.47 \tab 6.49 \tab 6.5 \tab 6.51 \tab 6.55 \tab 6.88 \tab 6.91 \tab 6.93 \tab 6.98 \tab 7.02 \tab 7.22 \tab 7.23 \tab 7.25 \tab 7.27 \tab 7.28 \tab 7.4 \tab 7.43 \tab 7.45 \tab 7.51 \tab 7.54 \tab 7.55 \tab 7.57 \tab 7.58 \tab 7.6 \tab 7.63 \tab 7.64 \tab 7.66 \tab 7.67 \tab 7.68 \tab 7.69 \tab 7.71 \tab 7.72 \tab 7.74 \tab 7.75 \tab 7.77 \tab 7.78 \tab 7.79 \tab 7.8 \tab 7.81 \tab 7.82 \tab 7.83 \tab 7.84 \tab 7.85 \tab 7.86 \tab 7.87 \tab 7.88 \tab 7.89 \tab 7.9 \tab 7.91 \tab 7.92 \tab 7.93 \tab 7.94 \tab 7.95 \tab 7.96 \tab 7.97 \tab 7.98 \tab 8 \tab iap \tab 0.73 \tab 1.44 \tab 2.3 \tab 2.44 \tab 2.47 \tab 2.75 \tab 3 \tab 3.24 \tab 3.42 \tab 4.17 \tab 4.23 \tab 4.33 \tab 5.74 \tab 5.94 \tab 6.07 \tab 6.16 \tab 6.22 \tab 6.29 \tab 6.37 \tab 6.44 \tab 6.65 \tab 6.79 \tab 6.82 \tab 7.12 \tab 7.16 \tab 7.49 \tab 7.59 \tab 7.62 \tab 7.65 \tab 7.7 \tab 7.73 \tab 7.76 \tab 0.9 \tab 2.2 \tab 2.35 \tab 2.39 \tab 3.66 \tab 3.83 \tab 4.35 \tab 4.7 \tab 4.93 \tab 5.06 \tab 5.27 \tab 5.49 \tab 5.63 \tab 5.97 \tab 6.99 \tab 7 \tab 7.21 \tab 2.59 \tab 2.62 \tab 6.8 \tab 7.04 \tab 2.29 \tab 4.4 \tab 4.99 \tab 5.37 \tab 6.66 \tab 7.05 \tab 7.07 \tab 2.09 \tab 2.51 \tab 5.28 \tab 6.94 \tab 7.24 \tab 5 \tab 6.32 \tab 2.64 \tab 3.48 \tab 4.73 \tab 5.8 \tab 6.08 \tab 7.03 \tab 7.5 \tab 1.68 \tab 3.51 \tab 4.08 \tab 5.23 \tab 7.08 \tab 7.47 \tab 4.1 \tab 5.75 \tab 3.99 \tab 4.31 \tab 6.25 \tab 6.63 \tab Total \cr
   1972 \tab 2 \tab 10 \tab 3 \tab 3 \tab 7 \tab 2 \tab 1 \tab 3 \tab 1 \tab 5 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 26 \tab 12 \tab 3 \tab 2 \tab 9 \tab 1 \tab 2 \tab 4 \tab 1 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 6 \tab 1 \tab 3 \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 75 \tab 22 \tab 9 \tab 1 \tab 3 \tab 1 \tab 3 \tab 9 \tab 2 \tab 4 \tab 7 \tab 1 \tab 4 \tab 6 \tab 4 \tab 3 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 4 \tab 7 \tab 1 \tab 6 \tab 1 \tab 6 \tab 2 \tab 1 \tab 2 \tab 2 \tab 12 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 7 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab 5 \tab 3 \tab 3 \tab 1 \tab 11 \tab 6 \tab 21 \tab 2 \tab 3 \tab 3 \tab 3 \tab 7 \tab 8 \tab 7 \tab 19 \tab 15 \tab 1 \tab 1 \tab 42 \tab 16 \tab 51 \tab 12 \tab 6 \tab 20 \tab 1 \tab 20 \tab 8 \tab 6 \tab 16 \tab 33 \tab 20 \tab 370 \tab 17 \tab 171 \tab 276 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 2 \tab 12 \tab 2 \tab 3 \tab 5 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 6 \tab 2 \tab 2 \tab 2 \tab 6 \tab 3 \tab 3 \tab 20 \tab 11 \tab 2 \tab 6 \tab 9 \tab - \tab 2 \tab 11 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 5 \tab 5 \tab 1 \tab - \tab 6 \tab 7 \tab 2 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 7 \tab 1 \tab 3 \tab 82 \tab 14 \tab 7 \tab - \tab 2 \tab - \tab 2 \tab 11 \tab 6 \tab 1 \tab 9 \tab - \tab 2 \tab 7 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 4 \tab 6 \tab - \tab 7 \tab 41 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 6 \tab 2 \tab - \tab 3 \tab 4 \tab - \tab 3 \tab - \tab 6 \tab - \tab 6 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab - \tab 4 \tab 4 \tab 10 \tab 8 \tab 20 \tab 12 \tab 4 \tab 6 \tab 6 \tab 8 \tab 6 \tab 6 \tab 17 \tab 13 \tab 2 \tab 2 \tab 49 \tab 6 \tab 25 \tab 13 \tab 1 \tab 21 \tab 1 \tab 12 \tab 11 \tab 4 \tab 17 \tab 27 \tab 23 \tab 313 \tab 24 \tab 169 \tab 191 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 4 \tab 2 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 7 \tab 5 \tab 2 \tab 3 \tab 2 \tab - \tab 7 \tab 1 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab 4 \tab 26 \tab 9 \tab 1 \tab 7 \tab 2 \tab 2 \tab 4 \tab 7 \tab 1 \tab 4 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 3 \tab 5 \tab 6 \tab - \tab 8 \tab 8 \tab 3 \tab 2 \tab 4 \tab - \tab - \tab - \tab 5 \tab 1 \tab 2 \tab 9 \tab 1 \tab - \tab 80 \tab 3 \tab 8 \tab 1 \tab 2 \tab 1 \tab - \tab 9 \tab 8 \tab 2 \tab 11 \tab - \tab 1 \tab 4 \tab 7 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 16 \tab 2 \tab 3 \tab - \tab 7 \tab 1 \tab 2 \tab 3 \tab 3 \tab 8 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 6 \tab 5 \tab - \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 6 \tab 4 \tab 19 \tab 8 \tab 4 \tab 2 \tab 5 \tab 7 \tab 5 \tab 6 \tab 18 \tab 16 \tab 4 \tab - \tab 51 \tab 7 \tab 37 \tab 8 \tab 2 \tab 19 \tab 2 \tab 15 \tab 10 \tab 8 \tab 18 \tab 17 \tab 19 \tab 317 \tab 17 \tab 187 \tab 215 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 2 \tab 8 \tab 2 \tab 1 \tab 7 \tab 3 \tab - \tab 4 \tab 3 \tab 2 \tab 1 \tab - \tab 10 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 23 \tab 12 \tab 7 \tab 7 \tab 5 \tab 3 \tab 3 \tab 7 \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 10 \tab 4 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 71 \tab 29 \tab 9 \tab 1 \tab 3 \tab 2 \tab - \tab 10 \tab - \tab 3 \tab 7 \tab - \tab 1 \tab 12 \tab 4 \tab 2 \tab - \tab - \tab - \tab 7 \tab 2 \tab - \tab 24 \tab - \tab 3 \tab 6 \tab 7 \tab - \tab 1 \tab - \tab - \tab 11 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 5 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab 2 \tab - \tab 4 \tab 2 \tab 16 \tab 3 \tab 15 \tab 5 \tab 3 \tab 1 \tab 3 \tab 9 \tab 8 \tab 5 \tab 12 \tab 17 \tab 3 \tab 3 \tab 53 \tab 10 \tab 39 \tab 12 \tab 2 \tab 17 \tab - \tab 12 \tab 4 \tab 11 \tab 8 \tab 35 \tab 24 \tab 318 \tab 32 \tab 194 \tab 190 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 2 \tab 9 \tab 2 \tab 3 \tab 10 \tab 2 \tab - \tab 5 \tab 3 \tab 5 \tab 3 \tab - \tab 5 \tab 1 \tab 5 \tab 5 \tab 1 \tab 1 \tab 3 \tab 32 \tab 17 \tab - \tab 6 \tab 6 \tab - \tab 2 \tab 4 \tab 5 \tab 2 \tab 4 \tab 1 \tab - \tab - \tab - \tab 2 \tab 9 \tab 6 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 2 \tab 60 \tab 28 \tab 11 \tab - \tab - \tab 2 \tab 1 \tab 8 \tab 1 \tab 6 \tab 9 \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 5 \tab 19 \tab 1 \tab 3 \tab - \tab 5 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab 4 \tab 8 \tab 7 \tab 15 \tab 10 \tab 4 \tab 11 \tab 2 \tab 9 \tab 5 \tab 2 \tab 17 \tab 5 \tab 1 \tab 1 \tab 46 \tab 16 \tab 43 \tab 11 \tab 2 \tab 18 \tab 2 \tab 13 \tab 9 \tab 5 \tab 20 \tab 30 \tab 16 \tab 320 \tab 26 \tab 197 \tab 219 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1 \tab 9 \tab 3 \tab 10 \tab 4 \tab 5 \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab 10 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 35 \tab 11 \tab 4 \tab 3 \tab 5 \tab 3 \tab 2 \tab 7 \tab 4 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 61 \tab 21 \tab 9 \tab 1 \tab - \tab - \tab 1 \tab 11 \tab 4 \tab 1 \tab 9 \tab 1 \tab 2 \tab 7 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 12 \tab - \tab - \tab 1 \tab 7 \tab 2 \tab 1 \tab 3 \tab - \tab 4 \tab 2 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab 4 \tab 3 \tab - \tab - \tab 6 \tab 1 \tab 2 \tab 2 \tab 5 \tab 5 \tab 19 \tab 4 \tab 3 \tab 5 \tab 6 \tab 13 \tab 2 \tab 5 \tab 19 \tab 7 \tab 1 \tab - \tab 42 \tab 14 \tab 42 \tab 15 \tab 3 \tab 23 \tab 3 \tab 14 \tab 7 \tab 6 \tab 31 \tab 26 \tab 24 \tab 332 \tab 24 \tab 206 \tab 236 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1 \tab 6 \tab - \tab 8 \tab 7 \tab 2 \tab - \tab 5 \tab - \tab 5 \tab 5 \tab - \tab 5 \tab 1 \tab 5 \tab 1 \tab 6 \tab 1 \tab 1 \tab 34 \tab 5 \tab - \tab 6 \tab 5 \tab 3 \tab - \tab 4 \tab 4 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 69 \tab 35 \tab 9 \tab - \tab 1 \tab - \tab 4 \tab 10 \tab 2 \tab 4 \tab 10 \tab - \tab 3 \tab 5 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 20 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 2 \tab 4 \tab 4 \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 10 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 5 \tab 1 \tab 2 \tab 2 \tab 3 \tab 3 \tab 17 \tab 10 \tab 1 \tab 10 \tab 7 \tab 19 \tab 10 \tab 5 \tab 18 \tab 13 \tab 1 \tab 4 \tab 63 \tab 12 \tab 24 \tab 12 \tab 2 \tab 17 \tab 1 \tab 23 \tab 11 \tab 9 \tab 23 \tab 27 \tab 42 \tab 296 \tab 21 \tab 191 \tab 244 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 4 \tab 5 \tab - \tab 9 \tab 8 \tab 4 \tab - \tab 9 \tab 5 \tab 4 \tab 1 \tab - \tab 5 \tab 1 \tab 8 \tab 3 \tab 2 \tab 1 \tab 1 \tab 23 \tab 9 \tab - \tab 4 \tab 3 \tab - \tab 2 \tab 4 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 8 \tab 2 \tab 5 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab 3 \tab - \tab 4 \tab - \tab 2 \tab 1 \tab 4 \tab 3 \tab 35 \tab 26 \tab 10 \tab - \tab 1 \tab - \tab 5 \tab 11 \tab 2 \tab 5 \tab 9 \tab 1 \tab 1 \tab 7 \tab 5 \tab - \tab 2 \tab - \tab 2 \tab 4 \tab - \tab 3 \tab 14 \tab - \tab 5 \tab 2 \tab 3 \tab 5 \tab - \tab 4 \tab - \tab 9 \tab - \tab - \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 6 \tab - \tab 2 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab 6 \tab 2 \tab 9 \tab 3 \tab 17 \tab 5 \tab 3 \tab 2 \tab 3 \tab 13 \tab 7 \tab 3 \tab 9 \tab 14 \tab - \tab 1 \tab 47 \tab 14 \tab 27 \tab 13 \tab 5 \tab 26 \tab 2 \tab 11 \tab 12 \tab 8 \tab 14 \tab 25 \tab 16 \tab 271 \tab 20 \tab 173 \tab 260 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 17 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 5 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 4 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1 \tab 13 \tab 3 \tab 5 \tab 5 \tab 4 \tab - \tab 2 \tab - \tab 8 \tab 2 \tab 5 \tab 8 \tab 2 \tab 11 \tab 2 \tab 6 \tab - \tab 1 \tab 20 \tab 3 \tab 2 \tab 4 \tab 3 \tab 1 \tab 4 \tab 7 \tab 2 \tab 2 \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 5 \tab 8 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 6 \tab 1 \tab 3 \tab 21 \tab 48 \tab 11 \tab - \tab - \tab 1 \tab 2 \tab 6 \tab 1 \tab 2 \tab 16 \tab - \tab - \tab 12 \tab 11 \tab - \tab 1 \tab - \tab 14 \tab 2 \tab 2 \tab 4 \tab 12 \tab - \tab 4 \tab 1 \tab 5 \tab 1 \tab 2 \tab 5 \tab 2 \tab 8 \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 9 \tab - \tab 4 \tab - \tab 12 \tab 4 \tab 1 \tab - \tab 5 \tab 2 \tab 11 \tab - \tab 12 \tab 6 \tab 17 \tab 4 \tab 4 \tab 13 \tab 2 \tab 12 \tab 7 \tab 8 \tab 10 \tab 16 \tab 1 \tab 3 \tab 63 \tab 18 \tab 44 \tab 11 \tab 1 \tab 42 \tab 3 \tab 14 \tab 4 \tab 17 \tab 17 \tab 51 \tab 30 \tab 326 \tab 43 \tab 209 \tab 378 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 13 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab 3 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 4 \tab 12 \tab 6 \tab 10 \tab 8 \tab 2 \tab - \tab 3 \tab 10 \tab 1 \tab 6 \tab 4 \tab 13 \tab 1 \tab 4 \tab 2 \tab 5 \tab 3 \tab 1 \tab 40 \tab 17 \tab 2 \tab 7 \tab 7 \tab 1 \tab - \tab 8 \tab 4 \tab 3 \tab 2 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 9 \tab - \tab 1 \tab 1 \tab 6 \tab 1 \tab 35 \tab 41 \tab 7 \tab - \tab 1 \tab - \tab 7 \tab 10 \tab 3 \tab - \tab 6 \tab - \tab - \tab 7 \tab 9 \tab 1 \tab - \tab - \tab 8 \tab 1 \tab 2 \tab 3 \tab 16 \tab - \tab 1 \tab 2 \tab 4 \tab 1 \tab 4 \tab 3 \tab - \tab 10 \tab 1 \tab - \tab 4 \tab - \tab - \tab 3 \tab 1 \tab 5 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 5 \tab - \tab 2 \tab 3 \tab 15 \tab 6 \tab 23 \tab 14 \tab 4 \tab 3 \tab 4 \tab 16 \tab 8 \tab 7 \tab 10 \tab 13 \tab - \tab 2 \tab 58 \tab 24 \tab 38 \tab 13 \tab 7 \tab 42 \tab 3 \tab 12 \tab 9 \tab 5 \tab 18 \tab 35 \tab 37 \tab 243 \tab 24 \tab 171 \tab 238 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 7 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1 \tab 5 \tab 1 \tab 5 \tab 3 \tab 3 \tab - \tab 3 \tab 2 \tab 3 \tab 3 \tab - \tab 3 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 52 \tab 27 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab 2 \tab 2 \tab - \tab - \tab - \tab 101 \tab 32 \tab 12 \tab - \tab 2 \tab 3 \tab 6 \tab 8 \tab - \tab 2 \tab 7 \tab 1 \tab 1 \tab 7 \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 7 \tab 1 \tab 2 \tab 8 \tab 1 \tab 4 \tab 2 \tab 4 \tab 2 \tab 6 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 5 \tab - \tab 7 \tab 2 \tab 4 \tab - \tab 2 \tab 1 \tab 4 \tab - \tab 3 \tab 1 \tab 10 \tab 7 \tab 12 \tab 8 \tab 1 \tab - \tab 4 \tab 17 \tab 8 \tab 33 \tab 11 \tab 16 \tab - \tab - \tab 28 \tab 12 \tab 75 \tab 9 \tab 3 \tab 37 \tab 1 \tab 13 \tab 7 \tab 3 \tab 14 \tab 34 \tab 15 \tab 282 \tab 26 \tab 106 \tab 249 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 14 \tab 3 \tab 8 \tab 9 \tab 4 \tab - \tab 5 \tab 4 \tab 5 \tab 1 \tab 1 \tab 16 \tab 5 \tab 6 \tab 5 \tab 6 \tab 2 \tab 4 \tab 25 \tab 14 \tab 1 \tab 6 \tab 5 \tab 1 \tab 3 \tab 7 \tab 3 \tab 2 \tab 2 \tab 6 \tab - \tab 3 \tab 4 \tab 2 \tab 7 \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 3 \tab - \tab 2 \tab 6 \tab 4 \tab 3 \tab 55 \tab 29 \tab 14 \tab 2 \tab 2 \tab 2 \tab 2 \tab 12 \tab 2 \tab 1 \tab 15 \tab 1 \tab - \tab 4 \tab 4 \tab - \tab - \tab 2 \tab 4 \tab 2 \tab - \tab 2 \tab 4 \tab 1 \tab 2 \tab 2 \tab 6 \tab 2 \tab 3 \tab 4 \tab 1 \tab 5 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 6 \tab - \tab 5 \tab - \tab 5 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 21 \tab 6 \tab 17 \tab 5 \tab 3 \tab 11 \tab 5 \tab 17 \tab 7 \tab 3 \tab 19 \tab 13 \tab 2 \tab - \tab 42 \tab 15 \tab 31 \tab 6 \tab 6 \tab 35 \tab 2 \tab 10 \tab 15 \tab 10 \tab 24 \tab 44 \tab 24 \tab 274 \tab 30 \tab 131 \tab 219 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 5 \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 3 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 3 \tab 8 \tab 3 \tab 7 \tab 5 \tab 3 \tab 2 \tab 8 \tab 3 \tab 2 \tab 1 \tab 2 \tab 6 \tab 3 \tab 5 \tab 1 \tab 2 \tab 1 \tab 4 \tab 30 \tab 12 \tab - \tab 8 \tab 4 \tab 2 \tab - \tab 9 \tab 4 \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 8 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab 7 \tab - \tab 52 \tab 55 \tab 6 \tab - \tab 2 \tab 2 \tab - \tab 12 \tab 4 \tab 1 \tab 10 \tab - \tab 3 \tab 5 \tab 7 \tab - \tab - \tab 4 \tab 2 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 5 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab 5 \tab - \tab 9 \tab - \tab 2 \tab 1 \tab - \tab - \tab 7 \tab 1 \tab 2 \tab 3 \tab 13 \tab 5 \tab 15 \tab 5 \tab 2 \tab 7 \tab 2 \tab 15 \tab 9 \tab 3 \tab 6 \tab 12 \tab - \tab 4 \tab 41 \tab 12 \tab 19 \tab 11 \tab 2 \tab 19 \tab - \tab 7 \tab 11 \tab 6 \tab 12 \tab 32 \tab 24 \tab 288 \tab 12 \tab 143 \tab 258 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 17 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 11 \tab 2 \tab 9 \tab 11 \tab 2 \tab 1 \tab 2 \tab 5 \tab 10 \tab 2 \tab 9 \tab 5 \tab 1 \tab 7 \tab 2 \tab 5 \tab 2 \tab 2 \tab 32 \tab 9 \tab 1 \tab 6 \tab 6 \tab 1 \tab 3 \tab 3 \tab 4 \tab 1 \tab - \tab 1 \tab 6 \tab 3 \tab 2 \tab 2 \tab 5 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 5 \tab 32 \tab 35 \tab 10 \tab - \tab 2 \tab 1 \tab - \tab 9 \tab 3 \tab 9 \tab 13 \tab - \tab 1 \tab 6 \tab 5 \tab 2 \tab - \tab - \tab 5 \tab - \tab 2 \tab 1 \tab 20 \tab - \tab 6 \tab 2 \tab 2 \tab 4 \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 8 \tab - \tab 7 \tab 2 \tab 3 \tab 3 \tab - \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 12 \tab 15 \tab 25 \tab 7 \tab 5 \tab 9 \tab 3 \tab 16 \tab 12 \tab 12 \tab 7 \tab 14 \tab 1 \tab 2 \tab 51 \tab 20 \tab 35 \tab 11 \tab 2 \tab 31 \tab 3 \tab 13 \tab 11 \tab 3 \tab 16 \tab 46 \tab 26 \tab 343 \tab 44 \tab 205 \tab 341 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 8 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 4 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 10 \tab 3 \tab 4 \tab 7 \tab 9 \tab 2 \tab 3 \tab 6 \tab 4 \tab 3 \tab 1 \tab 3 \tab - \tab 7 \tab 1 \tab 5 \tab 4 \tab - \tab 38 \tab 7 \tab - \tab 4 \tab 5 \tab - \tab 1 \tab 8 \tab - \tab - \tab 3 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab 8 \tab 1 \tab - \tab 5 \tab 2 \tab 3 \tab 1 \tab 2 \tab 2 \tab 55 \tab 44 \tab 8 \tab - \tab 1 \tab 2 \tab 1 \tab 9 \tab 5 \tab 3 \tab 5 \tab - \tab 2 \tab 8 \tab 10 \tab 1 \tab - \tab - \tab 2 \tab 7 \tab 2 \tab - \tab 8 \tab 1 \tab 9 \tab 7 \tab - \tab 3 \tab 1 \tab 3 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 7 \tab - \tab 4 \tab 3 \tab 9 \tab 1 \tab - \tab 2 \tab 5 \tab 1 \tab 4 \tab 2 \tab 15 \tab 15 \tab 20 \tab 8 \tab 3 \tab - \tab 2 \tab 12 \tab 7 \tab 3 \tab 11 \tab 10 \tab - \tab 3 \tab 41 \tab 6 \tab 32 \tab 7 \tab 6 \tab 25 \tab 3 \tab 14 \tab 9 \tab 6 \tab 11 \tab 37 \tab 30 \tab 266 \tab 28 \tab 127 \tab 242 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 12 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 2 \tab 12 \tab - \tab 5 \tab 7 \tab 6 \tab 1 \tab 5 \tab 10 \tab 8 \tab 5 \tab 5 \tab 7 \tab 2 \tab 6 \tab 4 \tab 5 \tab - \tab 1 \tab 46 \tab 12 \tab 3 \tab - \tab 4 \tab 2 \tab 3 \tab 7 \tab 2 \tab 1 \tab 3 \tab 5 \tab 2 \tab 2 \tab - \tab 3 \tab 12 \tab - \tab 2 \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 2 \tab 1 \tab 5 \tab 1 \tab 66 \tab 29 \tab 5 \tab - \tab 1 \tab 1 \tab 4 \tab 11 \tab 2 \tab 6 \tab 8 \tab - \tab - \tab 6 \tab 7 \tab - \tab - \tab 1 \tab 5 \tab - \tab 6 \tab 2 \tab 8 \tab - \tab 2 \tab 1 \tab 6 \tab - \tab 2 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 6 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab 6 \tab 9 \tab 13 \tab 6 \tab 16 \tab 8 \tab 7 \tab 11 \tab 8 \tab 15 \tab 11 \tab 3 \tab 9 \tab 10 \tab - \tab - \tab 36 \tab 9 \tab 39 \tab 5 \tab 3 \tab 25 \tab 4 \tab 12 \tab 9 \tab 7 \tab 27 \tab 46 \tab 24 \tab 245 \tab 33 \tab 144 \tab 256 \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 5 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1537 \cr
   1990 \tab 2 \tab 8 \tab 2 \tab 1 \tab 6 \tab 1 \tab 2 \tab 9 \tab 5 \tab 5 \tab 3 \tab - \tab 8 \tab 2 \tab 8 \tab 2 \tab 1 \tab 1 \tab 3 \tab 37 \tab 9 \tab - \tab 2 \tab 4 \tab 1 \tab 4 \tab 7 \tab 4 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab 3 \tab 2 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 5 \tab - \tab 62 \tab 38 \tab 8 \tab 1 \tab 1 \tab - \tab 3 \tab 6 \tab 7 \tab 4 \tab 11 \tab - \tab - \tab 3 \tab 5 \tab - \tab - \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 8 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 8 \tab 1 \tab 5 \tab 3 \tab 14 \tab 8 \tab 12 \tab 4 \tab 7 \tab 6 \tab 6 \tab 10 \tab 9 \tab 3 \tab 6 \tab 9 \tab 1 \tab 3 \tab 36 \tab 12 \tab 27 \tab 10 \tab 1 \tab 19 \tab 3 \tab 14 \tab 7 \tab 10 \tab 18 \tab 39 \tab 25 \tab 214 \tab 29 \tab 118 \tab 248 \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 27 \tab 159 \tab 40 \tab 93 \tab 112 \tab 56 \tab 9 \tab 77 \tab 61 \tab 71 \tab 45 \tab 28 \tab 118 \tab 23 \tab 82 \tab 31 \tab 56 \tab 25 \tab 30 \tab 539 \tab 196 \tab 29 \tab 78 \tab 82 \tab 22 \tab 35 \tab 104 \tab 43 \tab 32 \tab 24 \tab 45 \tab 22 \tab 17 \tab 17 \tab 43 \tab 99 \tab 35 \tab 24 \tab 19 \tab 16 \tab 23 \tab 18 \tab 11 \tab 27 \tab 24 \tab 10 \tab 50 \tab 15 \tab 33 \tab 49 \tab 41 \tab 26 \tab 1012 \tab 529 \tab 153 \tab 7 \tab 24 \tab 18 \tab 41 \tab 162 \tab 52 \tab 54 \tab 162 \tab 6 \tab 22 \tab 110 \tab 93 \tab 14 \tab 9 \tab 11 \tab 55 \tab 41 \tab 30 \tab 34 \tab 202 \tab 8 \tab 60 \tab 76 \tab 73 \tab 32 \tab 23 \tab 37 \tab 22 \tab 99 \tab 31 \tab 1 \tab 22 \tab 13 \tab 4 \tab 36 \tab 9 \tab 92 \tab 3 \tab 60 \tab 19 \tab 58 \tab 37 \tab 4 \tab 16 \tab 68 \tab 17 \tab 61 \tab 43 \tab 193 \tab 113 \tab 300 \tab 119 \tab 61 \tab 100 \tab 71 \tab 215 \tab 129 \tab 114 \tab 218 \tab 213 \tab 18 \tab 29 \tab 789 \tab 223 \tab 628 \tab 179 \tab 54 \tab 436 \tab 34 \tab 229 \tab 154 \tab 124 \tab 304 \tab 584 \tab 419 \tab 5018 \tab 450 \tab 2842 \tab 50385 \tab 23 \tab 8 \tab 6 \tab 20 \tab 7 \tab 12 \tab 11 \tab 10 \tab 2 \tab 8 \tab 10 \tab 4 \tab 17 \tab 9 \tab 12 \tab 4 \tab 1 \tab 6 \tab 83 \tab 7 \tab 6 \tab 7 \tab 12 \tab 16 \tab 14 \tab 24 \tab 4 \tab 14 \tab 31 \tab 5 \tab 14 \tab 33 \tab 6 \tab 10 \tab 11 \tab 1 \tab 17 \tab 27 \tab 4 \tab 25 \tab 9 \tab 3 \tab 5 \tab 2 \tab 9 \tab 22 \tab 12 \tab 6 \tab 3 \tab 13 \tab 6 \tab 18 \tab 6 \tab 8 \tab 7 \tab 3 \tab 13 \tab 4 \tab 2 \tab 4 \tab 4 \tab 1 \tab 4 \tab 9 \tab 3 \tab 1 \tab 6 \tab 15 \tab 2 \tab 8 \tab 8 \tab 1 \tab 4 \tab 4 \tab 4 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
