% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_19.R
\name{padotpre}
\alias{padotpre}
\title{Fa. temme prestige scale}
\source{
General Social Survey https://gss.norc.org
}
\description{
padotpre
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 10.8 \tab 11.5 \tab 12.8 \tab 13.8 \tab 14.5 \tab 15.5 \tab 16.7 \tab 16.8 \tab 18.1 \tab 18.2 \tab 18.3 \tab 18.7 \tab 19 \tab 19.2 \tab 19.3 \tab 19.9 \tab 20.2 \tab 20.4 \tab 20.5 \tab 21 \tab 21.4 \tab 21.6 \tab 21.7 \tab 22.2 \tab 22.6 \tab 22.7 \tab 22.9 \tab 23 \tab 23.1 \tab 23.2 \tab 23.4 \tab 23.9 \tab 24.2 \tab 24.8 \tab 24.9 \tab 25.2 \tab 25.3 \tab 25.7 \tab 25.9 \tab 26 \tab 26.2 \tab 26.4 \tab 26.5 \tab 26.8 \tab 26.9 \tab 27.1 \tab 27.4 \tab 27.5 \tab 27.6 \tab 27.9 \tab 28 \tab 28.3 \tab 28.7 \tab 28.8 \tab 28.9 \tab 29.1 \tab 29.6 \tab 29.7 \tab 29.9 \tab 30.2 \tab 30.3 \tab 30.4 \tab 30.5 \tab 30.6 \tab 30.7 \tab 30.9 \tab 31 \tab 31.1 \tab 31.2 \tab 31.3 \tab 31.7 \tab 31.8 \tab 32.1 \tab 32.3 \tab 32.6 \tab 32.8 \tab 33.2 \tab 33.4 \tab 33.6 \tab 33.7 \tab 33.9 \tab 34 \tab 34.1 \tab 34.4 \tab 34.5 \tab 34.8 \tab 35.1 \tab 35.5 \tab 35.7 \tab 35.8 \tab 36.2 \tab 36.3 \tab 36.6 \tab 37.3 \tab 37.5 \tab 37.6 \tab 37.7 \tab 38.2 \tab 38.4 \tab 38.6 \tab 38.7 \tab 38.9 \tab 39.3 \tab 39.4 \tab 39.5 \tab 39.6 \tab 39.7 \tab 39.8 \tab 40.4 \tab 40.5 \tab 40.6 \tab 40.7 \tab 41 \tab 41.2 \tab 41.4 \tab 41.5 \tab 41.6 \tab 41.7 \tab 42.4 \tab 42.5 \tab 42.6 \tab 42.8 \tab 42.9 \tab 43 \tab 43.1 \tab 43.2 \tab 43.4 \tab 43.6 \tab 43.7 \tab 43.8 \tab 44.1 \tab 44.4 \tab 44.5 \tab 44.6 \tab 44.7 \tab 44.9 \tab 45 \tab 45.2 \tab 45.3 \tab 45.6 \tab 45.7 \tab 45.9 \tab 46.3 \tab 46.4 \tab 46.5 \tab 46.9 \tab 47.1 \tab 47.5 \tab 47.8 \tab 48 \tab 48.3 \tab 48.6 \tab 49 \tab 49.3 \tab 49.5 \tab 49.6 \tab 49.7 \tab 50.2 \tab 50.4 \tab 50.7 \tab 52.3 \tab 53.7 \tab 54.2 \tab 55.5 \tab 55.8 \tab 56.3 \tab 57.3 \tab 59.6 \tab 60.1 \tab 60.5 \tab 60.6 \tab 60.8 \tab 61.4 \tab 61.8 \tab 62 \tab 62.6 \tab 62.7 \tab 62.9 \tab 63.6 \tab 64.1 \tab 64.5 \tab 65.3 \tab 65.5 \tab 67.1 \tab 67.9 \tab 68.1 \tab 68.2 \tab 68.4 \tab 69 \tab 69.2 \tab 69.7 \tab 7.4 \tab 70.9 \tab 71.2 \tab 71.8 \tab 73 \tab 76.4 \tab 76.7 \tab 78 \tab 88.4 \tab 9.3 \tab iap \tab 12.6 \tab 15 \tab 15.3 \tab 15.8 \tab 20.9 \tab 23.3 \tab 23.6 \tab 24 \tab 24.1 \tab 24.3 \tab 24.4 \tab 25.1 \tab 25.6 \tab 26.3 \tab 26.6 \tab 26.7 \tab 27.3 \tab 29 \tab 29.2 \tab 29.3 \tab 29.8 \tab 32 \tab 32.2 \tab 32.4 \tab 33 \tab 35.2 \tab 35.6 \tab 36 \tab 37.9 \tab 38.5 \tab 38.8 \tab 39.1 \tab 41.3 \tab 42.1 \tab 44.2 \tab 45.4 \tab 45.8 \tab 46 \tab 46.8 \tab 47.7 \tab 48.1 \tab 48.8 \tab 49.4 \tab 51.2 \tab 52.2 \tab 52.5 \tab 53.5 \tab 53.8 \tab 53.9 \tab 55.7 \tab 56 \tab 56.4 \tab 58 \tab 59.7 \tab 60.7 \tab 66 \tab 67 \tab 68.8 \tab 70.4 \tab 72.8 \tab 8.5 \tab 0 \tab 1.6 \tab 11.4 \tab 14.1 \tab 14.6 \tab 17.8 \tab 23.5 \tab 25.5 \tab 28.4 \tab 30 \tab 32.7 \tab 33.5 \tab 35 \tab 35.3 \tab 35.4 \tab 36.7 \tab 42.7 \tab 47.6 \tab 49.1 \tab 5.3 \tab 51.9 \tab 54.6 \tab 58.3 \tab 62.5 \tab 64.6 \tab 67.4 \tab 68.9 \tab 73.6 \tab 73.7 \tab 18.4 \tab 21.3 \tab 22 \tab 30.8 \tab 39.2 \tab 43.3 \tab 58.5 \tab 60.9 \tab 75.3 \tab 2.1 \tab 27.2 \tab 28.5 \tab 45.5 \tab 47.9 \tab 56.5 \tab 59.5 \tab 61.7 \tab 70.3 \tab 71.1 \tab 72 \tab 17.2 \tab 25.8 \tab 43.5 \tab 48.4 \tab 71.9 \tab 73.2 \tab 24.7 \tab 28.6 \tab 33.3 \tab 34.9 \tab 46.7 \tab 46.1 \tab 52.4 \tab 52.8 \tab 61.3 \tab 64.7 \tab 72.1 \tab 74.7 \tab 19.8 \tab 46.2 \tab 47.2 \tab 68 \tab 71 \tab 74.9 \tab 37.4 \tab 38.1 \tab 69.3 \tab 17.4 \tab 35.9 \tab 39.9 \tab 50.9 \tab 73.3 \tab 20.6 \tab 56.1 \tab 65.2 \tab 67.6 \tab 74.2 \tab 25 \tab 71.4 \tab 51.8 \tab 55.3 \tab 50.3 \tab Total \cr
   1972 \tab 40 \tab 1 \tab 1 \tab 4 \tab 8 \tab 6 \tab 1 \tab 13 \tab 1 \tab 3 \tab 4 \tab 2 \tab 1 \tab 4 \tab 2 \tab 5 \tab 2 \tab 6 \tab 9 \tab 2 \tab 1 \tab 2 \tab 3 \tab 2 \tab 3 \tab 14 \tab 7 \tab 13 \tab 2 \tab 2 \tab 7 \tab 2 \tab 10 \tab 3 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 1 \tab 27 \tab 2 \tab 8 \tab 1 \tab 3 \tab 1 \tab 6 \tab 12 \tab 4 \tab 6 \tab 2 \tab 3 \tab 2 \tab 29 \tab 3 \tab 1 \tab 5 \tab 9 \tab 5 \tab 2 \tab 7 \tab 1 \tab 1 \tab 310 \tab 3 \tab 2 \tab 9 \tab 12 \tab 9 \tab 7 \tab 2 \tab 6 \tab 4 \tab 2 \tab 6 \tab 5 \tab 4 \tab 2 \tab 7 \tab 1 \tab 10 \tab 7 \tab 5 \tab 1 \tab 2 \tab 15 \tab 9 \tab 2 \tab 5 \tab 1 \tab 17 \tab 9 \tab 23 \tab 1 \tab 4 \tab 5 \tab 2 \tab 3 \tab 36 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 17 \tab 2 \tab 24 \tab 3 \tab 1 \tab 12 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 18 \tab 5 \tab 7 \tab 6 \tab 4 \tab 6 \tab 1 \tab 12 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 7 \tab 2 \tab 8 \tab 1 \tab 10 \tab 1 \tab 1 \tab 5 \tab 6 \tab 6 \tab 1 \tab 77 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 11 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 4 \tab 4 \tab 10 \tab 31 \tab 2 \tab 9 \tab 3 \tab 3 \tab 2 \tab 1 \tab 2 \tab 7 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 7 \tab 1 \tab 276 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 33 \tab 1 \tab 2 \tab 3 \tab 11 \tab 8 \tab - \tab 9 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 12 \tab 17 \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 5 \tab 3 \tab 6 \tab - \tab 4 \tab - \tab 3 \tab 11 \tab 6 \tab - \tab 1 \tab 6 \tab 5 \tab 1 \tab 7 \tab 2 \tab 4 \tab 34 \tab 1 \tab 8 \tab 1 \tab 4 \tab - \tab 6 \tab 6 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 25 \tab 2 \tab - \tab 2 \tab 8 \tab 6 \tab 6 \tab 6 \tab - \tab 2 \tab 253 \tab 4 \tab - \tab 6 \tab 13 \tab 4 \tab 8 \tab - \tab 6 \tab 2 \tab - \tab 3 \tab 9 \tab 6 \tab 1 \tab 2 \tab - \tab 3 \tab 3 \tab 5 \tab 3 \tab 1 \tab 11 \tab 8 \tab - \tab 6 \tab 1 \tab 10 \tab 11 \tab 15 \tab 1 \tab 1 \tab 4 \tab 2 \tab 3 \tab 39 \tab 6 \tab - \tab 1 \tab 2 \tab 3 \tab 5 \tab 9 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 13 \tab - \tab 19 \tab - \tab 1 \tab 11 \tab - \tab - \tab 11 \tab 2 \tab 4 \tab - \tab 6 \tab 26 \tab 6 \tab 10 \tab 8 \tab 3 \tab 3 \tab 1 \tab 11 \tab 1 \tab 1 \tab - \tab 6 \tab - \tab 7 \tab 3 \tab 7 \tab 3 \tab 12 \tab 1 \tab 2 \tab 6 \tab 6 \tab 5 \tab 2 \tab 83 \tab - \tab 4 \tab - \tab 1 \tab - \tab 7 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 4 \tab 12 \tab 25 \tab - \tab 15 \tab 1 \tab 7 \tab 1 \tab - \tab 5 \tab 2 \tab - \tab 2 \tab 4 \tab 1 \tab 4 \tab - \tab 5 \tab 5 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab 3 \tab 2 \tab - \tab 5 \tab 1 \tab 191 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 3 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 41 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 37 \tab 2 \tab 4 \tab 1 \tab 7 \tab 5 \tab - \tab 3 \tab - \tab 7 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab 7 \tab 10 \tab - \tab - \tab 7 \tab 1 \tab 1 \tab 2 \tab 7 \tab 7 \tab 10 \tab 1 \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab 6 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 2 \tab 22 \tab 1 \tab 7 \tab 3 \tab 5 \tab - \tab 12 \tab 8 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 16 \tab 5 \tab 1 \tab 5 \tab 8 \tab 9 \tab 1 \tab 16 \tab 1 \tab 8 \tab 264 \tab - \tab - \tab 7 \tab 8 \tab 4 \tab 10 \tab - \tab 7 \tab 4 \tab - \tab 2 \tab 7 \tab 10 \tab - \tab 5 \tab 1 \tab 7 \tab 6 \tab 12 \tab - \tab 1 \tab 10 \tab 4 \tab - \tab 4 \tab - \tab 20 \tab 10 \tab 8 \tab - \tab 4 \tab 16 \tab 3 \tab 2 \tab 44 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab 4 \tab 10 \tab 3 \tab - \tab - \tab - \tab 6 \tab 12 \tab 2 \tab 9 \tab - \tab - \tab 10 \tab - \tab - \tab 7 \tab - \tab 3 \tab - \tab 2 \tab 32 \tab 8 \tab 11 \tab 3 \tab 7 \tab 3 \tab - \tab 9 \tab 7 \tab 2 \tab 2 \tab 2 \tab 1 \tab 9 \tab 2 \tab 4 \tab - \tab 4 \tab 2 \tab - \tab 8 \tab 4 \tab 5 \tab 3 \tab 81 \tab 1 \tab 5 \tab - \tab 2 \tab - \tab 8 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 3 \tab 7 \tab 28 \tab - \tab 13 \tab 7 \tab 2 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 2 \tab 5 \tab 2 \tab 3 \tab 6 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 5 \tab - \tab 3 \tab - \tab 215 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 4 \tab 3 \tab 3 \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 15 \tab 3 \tab 3 \tab 2 \tab 5 \tab 4 \tab - \tab 2 \tab 2 \tab 3 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab 22 \tab 1 \tab 2 \tab 9 \tab 1 \tab 2 \tab 1 \tab 10 \tab 5 \tab 13 \tab - \tab - \tab 2 \tab 1 \tab 8 \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab 11 \tab 3 \tab 1 \tab 28 \tab 1 \tab 3 \tab 1 \tab 4 \tab - \tab 13 \tab 11 \tab 3 \tab 1 \tab 2 \tab 1 \tab 5 \tab 31 \tab 4 \tab 2 \tab 2 \tab 6 \tab 15 \tab 1 \tab 24 \tab - \tab - \tab 265 \tab 3 \tab 1 \tab 4 \tab 12 \tab 6 \tab 5 \tab 1 \tab 4 \tab 3 \tab - \tab 1 \tab 12 \tab 2 \tab - \tab - \tab 2 \tab 11 \tab 16 \tab 8 \tab 1 \tab 1 \tab 11 \tab 6 \tab 1 \tab 8 \tab 2 \tab 22 \tab 11 \tab 34 \tab - \tab 2 \tab 8 \tab 3 \tab 7 \tab 47 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab 14 \tab 2 \tab 24 \tab - \tab 3 \tab 6 \tab - \tab 1 \tab 4 \tab - \tab 4 \tab - \tab 11 \tab 24 \tab 2 \tab 8 \tab 9 \tab - \tab 4 \tab 1 \tab 12 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 3 \tab 6 \tab 1 \tab 6 \tab 1 \tab - \tab - \tab 6 \tab 10 \tab 2 \tab 71 \tab 3 \tab 5 \tab - \tab 2 \tab - \tab 9 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 8 \tab 30 \tab 3 \tab 9 \tab 4 \tab 5 \tab 1 \tab - \tab 3 \tab 4 \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 8 \tab 6 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab - \tab 7 \tab - \tab 190 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 6 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 17 \tab 2 \tab 1 \tab 4 \tab 5 \tab 9 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 4 \tab - \tab 2 \tab - \tab 16 \tab 19 \tab 2 \tab 1 \tab 5 \tab 1 \tab 3 \tab 4 \tab 16 \tab 6 \tab 18 \tab - \tab - \tab 2 \tab 2 \tab 6 \tab - \tab - \tab 6 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 24 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 8 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 27 \tab 4 \tab 1 \tab 2 \tab 8 \tab 16 \tab 11 \tab 19 \tab 1 \tab - \tab 282 \tab 2 \tab - \tab 7 \tab 13 \tab 9 \tab 12 \tab - \tab 5 \tab 4 \tab 2 \tab 3 \tab 5 \tab 6 \tab 2 \tab 2 \tab 1 \tab 4 \tab 7 \tab 4 \tab - \tab - \tab 8 \tab 9 \tab - \tab 6 \tab - \tab 20 \tab 8 \tab 29 \tab - \tab 4 \tab 4 \tab 1 \tab 1 \tab 35 \tab 3 \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 11 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 9 \tab 2 \tab 13 \tab 3 \tab 2 \tab 17 \tab - \tab 1 \tab 4 \tab 3 \tab 6 \tab - \tab 3 \tab 11 \tab 6 \tab 9 \tab 9 \tab 3 \tab 4 \tab - \tab 17 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 11 \tab 2 \tab 6 \tab - \tab 5 \tab - \tab 5 \tab 9 \tab 5 \tab 60 \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 11 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 7 \tab 2 \tab 9 \tab 35 \tab - \tab 8 \tab 5 \tab - \tab - \tab 3 \tab 3 \tab 5 \tab - \tab 3 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 10 \tab - \tab 219 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab 5 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 4 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 39 \tab 1 \tab 1 \tab 1 \tab 5 \tab 9 \tab - \tab 2 \tab 2 \tab 3 \tab 3 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 5 \tab 19 \tab 1 \tab 1 \tab 4 \tab 2 \tab 2 \tab 1 \tab 14 \tab 4 \tab 14 \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab 3 \tab - \tab 4 \tab 1 \tab - \tab 3 \tab 6 \tab - \tab 1 \tab 44 \tab - \tab 4 \tab 4 \tab 5 \tab - \tab 13 \tab 4 \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab 25 \tab 7 \tab 2 \tab 5 \tab 7 \tab 11 \tab 3 \tab 12 \tab - \tab - \tab 257 \tab 2 \tab - \tab 7 \tab 16 \tab 7 \tab 7 \tab 1 \tab 13 \tab 3 \tab 1 \tab 1 \tab 19 \tab 3 \tab 1 \tab 8 \tab 1 \tab 12 \tab 9 \tab 10 \tab 3 \tab 2 \tab 10 \tab 11 \tab - \tab 2 \tab - \tab 18 \tab 11 \tab 23 \tab - \tab 2 \tab 9 \tab 1 \tab 2 \tab 30 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 19 \tab 3 \tab - \tab - \tab - \tab 4 \tab 13 \tab - \tab 16 \tab 1 \tab 2 \tab 14 \tab - \tab - \tab 6 \tab - \tab 3 \tab - \tab 4 \tab 12 \tab 11 \tab 10 \tab 13 \tab 1 \tab 4 \tab - \tab 11 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 2 \tab 8 \tab - \tab 8 \tab - \tab - \tab - \tab 6 \tab 1 \tab 1 \tab 63 \tab 4 \tab 1 \tab 1 \tab 3 \tab - \tab 10 \tab 3 \tab 6 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 9 \tab 43 \tab 1 \tab 9 \tab 3 \tab 2 \tab - \tab 1 \tab 5 \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab 3 \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 5 \tab - \tab - \tab 10 \tab 3 \tab - \tab 4 \tab - \tab 236 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 5 \tab 3 \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 29 \tab 1 \tab 1 \tab 2 \tab 11 \tab 4 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 5 \tab 14 \tab - \tab - \tab 3 \tab 3 \tab 2 \tab 5 \tab 12 \tab 3 \tab 9 \tab - \tab 2 \tab 4 \tab - \tab 7 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab 4 \tab 2 \tab 43 \tab 2 \tab 4 \tab 3 \tab 3 \tab 2 \tab 12 \tab 4 \tab 5 \tab 2 \tab 1 \tab - \tab 1 \tab 26 \tab 8 \tab 3 \tab 2 \tab 4 \tab 11 \tab 9 \tab 20 \tab 1 \tab - \tab 236 \tab 4 \tab - \tab 5 \tab 12 \tab 6 \tab 6 \tab 1 \tab 6 \tab 3 \tab - \tab 2 \tab 11 \tab 6 \tab - \tab 8 \tab 2 \tab 7 \tab 11 \tab 14 \tab 1 \tab 3 \tab 10 \tab 3 \tab - \tab 2 \tab 1 \tab 12 \tab 10 \tab 38 \tab 1 \tab 3 \tab 8 \tab 1 \tab 2 \tab 55 \tab 1 \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 17 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 13 \tab 3 \tab 13 \tab 1 \tab - \tab 9 \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab 4 \tab 21 \tab 6 \tab 9 \tab 18 \tab 2 \tab 12 \tab 2 \tab 5 \tab 2 \tab 1 \tab 3 \tab 6 \tab - \tab 1 \tab 2 \tab 11 \tab 1 \tab 5 \tab - \tab 1 \tab - \tab 6 \tab 3 \tab 5 \tab 72 \tab 4 \tab 2 \tab - \tab - \tab - \tab 10 \tab 1 \tab 6 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab 6 \tab 39 \tab 1 \tab 12 \tab 5 \tab 2 \tab 1 \tab - \tab 5 \tab 5 \tab - \tab 5 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 9 \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 8 \tab - \tab - \tab 7 \tab - \tab 244 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 12 \tab 1 \tab - \tab 5 \tab 8 \tab 5 \tab - \tab 4 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 6 \tab 11 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 4 \tab 12 \tab 7 \tab 6 \tab - \tab 1 \tab 3 \tab 1 \tab 9 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 6 \tab 1 \tab 5 \tab 38 \tab 3 \tab 2 \tab - \tab 6 \tab 1 \tab 9 \tab 9 \tab 3 \tab 4 \tab 1 \tab - \tab 1 \tab 22 \tab 2 \tab 2 \tab - \tab 4 \tab 7 \tab 1 \tab 14 \tab - \tab - \tab 228 \tab - \tab - \tab 1 \tab 15 \tab 4 \tab 11 \tab - \tab 3 \tab 3 \tab 2 \tab 2 \tab 12 \tab 2 \tab 2 \tab 2 \tab - \tab 8 \tab 6 \tab 4 \tab 2 \tab - \tab 9 \tab 3 \tab - \tab 7 \tab - \tab 12 \tab 11 \tab 29 \tab - \tab 1 \tab 4 \tab 11 \tab 4 \tab 39 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 11 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 12 \tab 1 \tab 18 \tab - \tab 2 \tab 6 \tab - \tab - \tab 3 \tab 2 \tab 3 \tab - \tab 4 \tab 17 \tab 11 \tab 8 \tab 11 \tab 2 \tab 1 \tab 2 \tab 9 \tab 5 \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 11 \tab 3 \tab 7 \tab 3 \tab 3 \tab - \tab 4 \tab 8 \tab 8 \tab 40 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 13 \tab 2 \tab 2 \tab 5 \tab - \tab - \tab 3 \tab 4 \tab 1 \tab 1 \tab 5 \tab 5 \tab 26 \tab 1 \tab 11 \tab 9 \tab 2 \tab 2 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 2 \tab 5 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 9 \tab - \tab 2 \tab 8 \tab - \tab 260 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 4 \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 17 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 37 \tab 5 \tab 1 \tab 2 \tab 11 \tab 10 \tab - \tab 5 \tab 3 \tab 11 \tab 3 \tab 1 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab 13 \tab 34 \tab - \tab - \tab 10 \tab 1 \tab 1 \tab 1 \tab 17 \tab 7 \tab 17 \tab - \tab - \tab 2 \tab - \tab 10 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 11 \tab 3 \tab 3 \tab 44 \tab 4 \tab 8 \tab 2 \tab 2 \tab 2 \tab 8 \tab 8 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 48 \tab 4 \tab 2 \tab 3 \tab 16 \tab 14 \tab 11 \tab 12 \tab - \tab - \tab 269 \tab 3 \tab - \tab 5 \tab 12 \tab 7 \tab 8 \tab 1 \tab 10 \tab 3 \tab 1 \tab 7 \tab 10 \tab 8 \tab 2 \tab 5 \tab 1 \tab 9 \tab 7 \tab 12 \tab - \tab - \tab 21 \tab 5 \tab - \tab 7 \tab - \tab 20 \tab 6 \tab 49 \tab - \tab 2 \tab 5 \tab 9 \tab 2 \tab 48 \tab 5 \tab - \tab 1 \tab - \tab 14 \tab - \tab 9 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab 19 \tab 4 \tab 10 \tab 2 \tab 2 \tab 7 \tab - \tab - \tab 4 \tab 1 \tab 3 \tab - \tab 3 \tab 19 \tab 14 \tab 12 \tab 12 \tab 1 \tab 4 \tab 7 \tab 3 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 6 \tab 2 \tab 6 \tab 3 \tab 6 \tab - \tab 2 \tab 1 \tab 4 \tab 8 \tab 11 \tab 22 \tab 2 \tab 2 \tab 1 \tab 5 \tab - \tab 14 \tab 1 \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 13 \tab 26 \tab 3 \tab 5 \tab 2 \tab 2 \tab 5 \tab - \tab 6 \tab 8 \tab 1 \tab 2 \tab 8 \tab 1 \tab 3 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 4 \tab - \tab - \tab 5 \tab 3 \tab - \tab 5 \tab - \tab 378 \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 5 \tab 2 \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab 13 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 20 \tab 3 \tab - \tab 2 \tab 7 \tab 7 \tab - \tab 8 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 5 \tab 9 \tab 1 \tab 3 \tab 1 \tab 4 \tab - \tab 2 \tab 24 \tab 3 \tab 5 \tab 2 \tab - \tab 2 \tab 12 \tab 3 \tab 3 \tab - \tab - \tab 3 \tab - \tab - \tab 7 \tab 2 \tab 6 \tab 14 \tab 1 \tab 7 \tab - \tab 6 \tab - \tab 13 \tab 10 \tab 5 \tab 7 \tab 3 \tab - \tab 4 \tab 34 \tab 9 \tab 3 \tab 1 \tab 8 \tab 9 \tab 2 \tab 16 \tab - \tab - \tab 194 \tab 1 \tab - \tab 5 \tab 13 \tab 7 \tab 14 \tab - \tab 12 \tab 6 \tab - \tab - \tab 12 \tab 11 \tab 1 \tab 2 \tab 1 \tab 7 \tab 13 \tab 18 \tab 2 \tab - \tab 20 \tab 7 \tab - \tab 7 \tab - \tab 21 \tab 10 \tab 43 \tab - \tab 2 \tab 4 \tab 12 \tab 1 \tab 49 \tab 3 \tab 1 \tab - \tab 2 \tab 8 \tab - \tab 7 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 11 \tab 6 \tab 29 \tab - \tab 1 \tab 8 \tab - \tab 1 \tab 5 \tab - \tab 2 \tab - \tab 7 \tab 17 \tab 6 \tab 3 \tab 16 \tab 1 \tab 3 \tab 3 \tab 17 \tab - \tab 2 \tab 1 \tab 5 \tab - \tab 1 \tab 4 \tab 6 \tab 1 \tab 13 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab 4 \tab 36 \tab 4 \tab 2 \tab 1 \tab 4 \tab 1 \tab 8 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 10 \tab 12 \tab 52 \tab 3 \tab 14 \tab 3 \tab 2 \tab 4 \tab 1 \tab 6 \tab 2 \tab 1 \tab 6 \tab 2 \tab - \tab 1 \tab 3 \tab 10 \tab 6 \tab 3 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 10 \tab 6 \tab - \tab 8 \tab - \tab 238 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 4 \tab 7 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 6 \tab 3 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab - \tab - \tab 4 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 23 \tab 1 \tab - \tab 1 \tab 5 \tab 3 \tab - \tab 58 \tab 5 \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 20 \tab - \tab - \tab 3 \tab 1 \tab - \tab 7 \tab 9 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 6 \tab - \tab - \tab - \tab 1 \tab 31 \tab - \tab 6 \tab - \tab 26 \tab 6 \tab 4 \tab 3 \tab - \tab 2 \tab - \tab 33 \tab 7 \tab 1 \tab - \tab 13 \tab 8 \tab - \tab 7 \tab - \tab - \tab 234 \tab 1 \tab - \tab 4 \tab 9 \tab 8 \tab 5 \tab 1 \tab 4 \tab 1 \tab - \tab 2 \tab 9 \tab 8 \tab 1 \tab 1 \tab 1 \tab 4 \tab 2 \tab 7 \tab - \tab 1 \tab 12 \tab 4 \tab - \tab 4 \tab 1 \tab 9 \tab 8 \tab 36 \tab 2 \tab - \tab 1 \tab 19 \tab - \tab 21 \tab 7 \tab - \tab 2 \tab 1 \tab - \tab - \tab 10 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 7 \tab - \tab 10 \tab - \tab - \tab 5 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 6 \tab 16 \tab 6 \tab 7 \tab 15 \tab 2 \tab 2 \tab - \tab 27 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab 5 \tab 105 \tab - \tab 2 \tab - \tab 1 \tab - \tab 12 \tab - \tab 6 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab 4 \tab 1 \tab 6 \tab 5 \tab 55 \tab - \tab 5 \tab 3 \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab - \tab 3 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 8 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 3 \tab 1 \tab 249 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 8 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 23 \tab 2 \tab 2 \tab 3 \tab 1 \tab 5 \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 6 \tab - \tab - \tab 2 \tab 4 \tab 23 \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab 12 \tab 4 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab 3 \tab 4 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab 4 \tab - \tab 2 \tab 14 \tab - \tab 4 \tab - \tab 4 \tab - \tab 1 \tab 5 \tab 9 \tab 6 \tab - \tab 2 \tab 2 \tab 38 \tab 3 \tab 2 \tab 2 \tab 2 \tab 7 \tab 11 \tab 4 \tab - \tab - \tab 225 \tab 2 \tab - \tab 9 \tab 7 \tab 2 \tab 5 \tab 1 \tab 5 \tab 2 \tab 1 \tab 2 \tab 12 \tab 4 \tab 3 \tab 9 \tab - \tab 9 \tab 10 \tab 9 \tab 1 \tab - \tab 20 \tab 3 \tab - \tab 6 \tab - \tab 22 \tab 13 \tab 29 \tab - \tab 2 \tab 6 \tab 6 \tab 2 \tab 32 \tab 2 \tab 1 \tab - \tab - \tab 4 \tab - \tab 12 \tab - \tab - \tab 1 \tab - \tab 4 \tab 16 \tab 3 \tab 15 \tab 1 \tab 6 \tab 5 \tab - \tab - \tab 4 \tab 2 \tab 7 \tab - \tab 3 \tab 19 \tab 7 \tab 15 \tab 16 \tab 4 \tab 9 \tab 5 \tab 14 \tab 1 \tab 4 \tab 1 \tab 6 \tab - \tab 6 \tab 3 \tab 8 \tab 3 \tab 9 \tab - \tab 5 \tab - \tab 7 \tab 7 \tab 6 \tab 57 \tab 3 \tab 3 \tab 3 \tab 6 \tab - \tab 18 \tab 1 \tab - \tab - \tab 2 \tab - \tab 4 \tab - \tab - \tab 2 \tab 7 \tab 14 \tab 35 \tab 3 \tab 9 \tab 5 \tab 2 \tab 1 \tab 1 \tab 6 \tab 6 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 10 \tab 5 \tab 6 \tab 1 \tab 1 \tab 3 \tab - \tab 5 \tab 2 \tab - \tab 4 \tab - \tab - \tab 8 \tab 3 \tab - \tab 9 \tab 1 \tab 219 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 6 \tab - \tab - \tab 2 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 5 \tab 2 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 4 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 3 \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 23 \tab 1 \tab - \tab 2 \tab 3 \tab 6 \tab 1 \tab 5 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 4 \tab 10 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 6 \tab 11 \tab 9 \tab 4 \tab 1 \tab - \tab 2 \tab 7 \tab 8 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 5 \tab 4 \tab 1 \tab 10 \tab - \tab 5 \tab 1 \tab 4 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab 32 \tab 4 \tab 2 \tab 1 \tab 4 \tab 5 \tab 6 \tab 1 \tab - \tab - \tab 217 \tab 4 \tab - \tab 2 \tab 13 \tab 7 \tab 9 \tab - \tab 10 \tab 2 \tab 2 \tab 1 \tab 18 \tab 5 \tab 1 \tab 3 \tab - \tab 5 \tab 5 \tab 9 \tab 2 \tab 3 \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab 8 \tab 12 \tab 56 \tab 1 \tab 4 \tab 5 \tab 6 \tab 3 \tab 33 \tab 5 \tab - \tab 1 \tab - \tab 2 \tab - \tab 7 \tab 3 \tab - \tab - \tab - \tab 2 \tab 9 \tab 2 \tab 27 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab 5 \tab - \tab 1 \tab - \tab 2 \tab 18 \tab 8 \tab 5 \tab 15 \tab 6 \tab 5 \tab 8 \tab 12 \tab - \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 7 \tab - \tab 13 \tab 1 \tab 1 \tab 2 \tab 5 \tab 8 \tab 5 \tab 53 \tab 4 \tab 2 \tab 1 \tab 3 \tab 1 \tab 8 \tab 2 \tab 5 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab 5 \tab 2 \tab 8 \tab 33 \tab 4 \tab 4 \tab 10 \tab 3 \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 4 \tab 8 \tab 3 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 7 \tab 3 \tab 2 \tab 5 \tab - \tab 258 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 6 \tab - \tab 1 \tab 6 \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab 17 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab 6 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 5 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 38 \tab 3 \tab 1 \tab - \tab 6 \tab 6 \tab 2 \tab 7 \tab 2 \tab 3 \tab 1 \tab - \tab 3 \tab 3 \tab 2 \tab 4 \tab 6 \tab 11 \tab 42 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 4 \tab 14 \tab 7 \tab 5 \tab - \tab 4 \tab 3 \tab 3 \tab 7 \tab - \tab - \tab 2 \tab 5 \tab 2 \tab - \tab 6 \tab 2 \tab 1 \tab 27 \tab 1 \tab 12 \tab 2 \tab 1 \tab 1 \tab 4 \tab 2 \tab 8 \tab 2 \tab - \tab 1 \tab 1 \tab 41 \tab 1 \tab 3 \tab 5 \tab 13 \tab 10 \tab 9 \tab 20 \tab - \tab - \tab 269 \tab 2 \tab 1 \tab 2 \tab 14 \tab 4 \tab 10 \tab - \tab 11 \tab 2 \tab - \tab 7 \tab 19 \tab 6 \tab 1 \tab 3 \tab 1 \tab 9 \tab 3 \tab 7 \tab 5 \tab 1 \tab 12 \tab 7 \tab 1 \tab 4 \tab 1 \tab 13 \tab 10 \tab 35 \tab - \tab 6 \tab 7 \tab 11 \tab - \tab 44 \tab 7 \tab 1 \tab 1 \tab 2 \tab 5 \tab - \tab 6 \tab - \tab - \tab 4 \tab 2 \tab 3 \tab 15 \tab 2 \tab 9 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 6 \tab 19 \tab 5 \tab 6 \tab 14 \tab 2 \tab 6 \tab 8 \tab 9 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 2 \tab 3 \tab 13 \tab 5 \tab 13 \tab 1 \tab 2 \tab - \tab 1 \tab 5 \tab 7 \tab 36 \tab 4 \tab 3 \tab 3 \tab 1 \tab - \tab 13 \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 1 \tab 11 \tab 37 \tab 2 \tab 19 \tab 2 \tab 2 \tab 9 \tab 1 \tab 3 \tab 10 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 7 \tab 11 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 9 \tab 2 \tab - \tab 11 \tab 1 \tab 341 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 7 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 8 \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 8 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 18 \tab 5 \tab - \tab 2 \tab 5 \tab 12 \tab - \tab 7 \tab 2 \tab 10 \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab 18 \tab - \tab 1 \tab - \tab 2 \tab - \tab 5 \tab 3 \tab 9 \tab 4 \tab - \tab - \tab - \tab 2 \tab 9 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 8 \tab 3 \tab 1 \tab 13 \tab - \tab 3 \tab 3 \tab 6 \tab - \tab 4 \tab 4 \tab 4 \tab 5 \tab 2 \tab 2 \tab 1 \tab 35 \tab 4 \tab 3 \tab 3 \tab 15 \tab 8 \tab - \tab 8 \tab - \tab - \tab 203 \tab - \tab - \tab 6 \tab 8 \tab 6 \tab 2 \tab - \tab 9 \tab - \tab 2 \tab 5 \tab 12 \tab 6 \tab 2 \tab 5 \tab 3 \tab 6 \tab 2 \tab 10 \tab 2 \tab 2 \tab 10 \tab 1 \tab - \tab 3 \tab - \tab 20 \tab 10 \tab 46 \tab 1 \tab 2 \tab 8 \tab 2 \tab 7 \tab 30 \tab 4 \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 4 \tab 10 \tab 3 \tab 31 \tab 1 \tab - \tab 6 \tab - \tab - \tab 6 \tab - \tab 1 \tab - \tab 3 \tab 13 \tab 6 \tab 2 \tab 12 \tab - \tab 7 \tab 3 \tab 7 \tab - \tab 1 \tab - \tab 5 \tab 1 \tab 1 \tab 4 \tab 11 \tab 2 \tab 8 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 7 \tab 58 \tab - \tab 2 \tab 2 \tab 4 \tab 2 \tab 9 \tab 8 \tab 6 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 4 \tab 3 \tab 10 \tab 41 \tab 2 \tab 11 \tab 4 \tab 1 \tab 1 \tab - \tab 6 \tab 4 \tab - \tab 3 \tab 1 \tab - \tab 5 \tab - \tab 7 \tab 7 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 5 \tab - \tab 5 \tab 9 \tab 1 \tab 2 \tab 4 \tab 3 \tab - \tab 7 \tab - \tab 242 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 5 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 5 \tab 5 \tab 3 \tab 1 \tab 1 \tab 12 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab 11 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 5 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 16 \tab 2 \tab - \tab 1 \tab 3 \tab 6 \tab - \tab 5 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 6 \tab 29 \tab 1 \tab - \tab - \tab 3 \tab - \tab 2 \tab 9 \tab 3 \tab 9 \tab 1 \tab 2 \tab 5 \tab - \tab 2 \tab 7 \tab 1 \tab - \tab - \tab 1 \tab - \tab 6 \tab - \tab 1 \tab 23 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 4 \tab 5 \tab 5 \tab 3 \tab - \tab - \tab 2 \tab 45 \tab 3 \tab 1 \tab - \tab 9 \tab 14 \tab 8 \tab 8 \tab - \tab - \tab 202 \tab 1 \tab - \tab 8 \tab 5 \tab 4 \tab 3 \tab - \tab 13 \tab 1 \tab - \tab 4 \tab 10 \tab 5 \tab 2 \tab 2 \tab 1 \tab 9 \tab 4 \tab 13 \tab - \tab - \tab 10 \tab 7 \tab - \tab 6 \tab - \tab 24 \tab 11 \tab 31 \tab - \tab 1 \tab 6 \tab 2 \tab - \tab 25 \tab 4 \tab - \tab 1 \tab 6 \tab 5 \tab - \tab 13 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 15 \tab - \tab 17 \tab 2 \tab - \tab 9 \tab - \tab 1 \tab 5 \tab 1 \tab 5 \tab - \tab 7 \tab 10 \tab 5 \tab 8 \tab 15 \tab 2 \tab 6 \tab 2 \tab 12 \tab 1 \tab 2 \tab 2 \tab 5 \tab 3 \tab 1 \tab - \tab 8 \tab 1 \tab 9 \tab 2 \tab 4 \tab 1 \tab 4 \tab 12 \tab 7 \tab 66 \tab 2 \tab 4 \tab 2 \tab 5 \tab - \tab 6 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 9 \tab 7 \tab 12 \tab 53 \tab 2 \tab 10 \tab 5 \tab 3 \tab 6 \tab 1 \tab 9 \tab 9 \tab 3 \tab 5 \tab 2 \tab 1 \tab 4 \tab 3 \tab 6 \tab 7 \tab 4 \tab - \tab 2 \tab 4 \tab - \tab 3 \tab 2 \tab 2 \tab 6 \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 7 \tab - \tab 256 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 4 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 5 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 2 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1537 \cr
   1990 \tab 17 \tab - \tab 1 \tab - \tab 2 \tab 7 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 5 \tab 1 \tab - \tab 1 \tab 5 \tab 26 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 11 \tab 2 \tab 4 \tab - \tab 1 \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 3 \tab 1 \tab 9 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 4 \tab 1 \tab - \tab - \tab 3 \tab 37 \tab 3 \tab 2 \tab 2 \tab 8 \tab 11 \tab 5 \tab 8 \tab - \tab - \tab 164 \tab 1 \tab - \tab 6 \tab 12 \tab 1 \tab 12 \tab 1 \tab 10 \tab 1 \tab - \tab 2 \tab 20 \tab 5 \tab - \tab 3 \tab 2 \tab 2 \tab 11 \tab 9 \tab - \tab - \tab 6 \tab 4 \tab - \tab 3 \tab - \tab 20 \tab 6 \tab 40 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 30 \tab 2 \tab - \tab - \tab 2 \tab 3 \tab - \tab 9 \tab 3 \tab 1 \tab - \tab - \tab 4 \tab 13 \tab 2 \tab 13 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab 12 \tab 4 \tab 6 \tab 8 \tab 3 \tab 9 \tab 8 \tab 9 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 14 \tab - \tab 11 \tab 3 \tab 2 \tab 1 \tab 4 \tab 5 \tab 8 \tab 64 \tab 4 \tab 2 \tab 1 \tab 5 \tab - \tab 10 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 3 \tab 8 \tab 42 \tab 4 \tab 9 \tab 10 \tab - \tab 1 \tab 1 \tab 8 \tab 6 \tab 1 \tab 3 \tab 3 \tab - \tab 4 \tab - \tab 10 \tab 9 \tab 5 \tab 1 \tab - \tab 5 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 6 \tab - \tab 248 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 4 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 5 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 437 \tab 34 \tab 18 \tab 35 \tab 103 \tab 112 \tab 4 \tab 135 \tab 26 \tab 60 \tab 27 \tab 12 \tab 14 \tab 52 \tab 6 \tab 18 \tab 19 \tab 110 \tab 332 \tab 9 \tab 13 \tab 53 \tab 31 \tab 14 \tab 54 \tab 200 \tab 88 \tab 143 \tab 7 \tab 17 \tab 41 \tab 38 \tab 105 \tab 40 \tab 7 \tab 35 \tab 33 \tab 16 \tab 9 \tab 102 \tab 29 \tab 33 \tab 414 \tab 20 \tab 113 \tab 25 \tab 65 \tab 9 \tab 144 \tab 99 \tab 66 \tab 49 \tab 14 \tab 22 \tab 30 \tab 544 \tab 73 \tab 31 \tab 40 \tab 142 \tab 166 \tab 86 \tab 202 \tab 4 \tab 11 \tab 4072 \tab 33 \tab 4 \tab 93 \tab 194 \tab 95 \tab 134 \tab 9 \tab 134 \tab 44 \tab 13 \tab 50 \tab 202 \tab 97 \tab 21 \tab 67 \tab 18 \tab 122 \tab 122 \tab 156 \tab 23 \tab 17 \tab 198 \tab 92 \tab 4 \tab 83 \tab 8 \tab 288 \tab 167 \tab 564 \tab 8 \tab 42 \tab 102 \tab 92 \tab 41 \tab 637 \tab 57 \tab 4 \tab 20 \tab 28 \tab 55 \tab 11 \tab 161 \tab 23 \tab 13 \tab 13 \tab 11 \tab 60 \tab 218 \tab 34 \tab 297 \tab 17 \tab 26 \tab 135 \tab 2 \tab 6 \tab 73 \tab 14 \tab 50 \tab 1 \tab 74 \tab 304 \tab 116 \tab 136 \tab 200 \tab 43 \tab 88 \tab 51 \tab 196 \tab 31 \tab 27 \tab 22 \tab 56 \tab 16 \tab 49 \tab 36 \tab 145 \tab 26 \tab 140 \tab 18 \tab 31 \tab 24 \tab 72 \tab 99 \tab 87 \tab 1044 \tab 43 \tab 52 \tab 17 \tab 45 \tab 6 \tab 177 \tab 27 \tab 51 \tab 24 \tab 18 \tab 15 \tab 26 \tab 20 \tab 21 \tab 46 \tab 64 \tab 159 \tab 631 \tab 31 \tab 172 \tab 81 \tab 39 \tab 37 \tab 13 \tab 81 \tab 81 \tab 11 \tab 45 \tab 43 \tab 9 \tab 41 \tab 23 \tab 88 \tab 104 \tab 49 \tab 8 \tab 6 \tab 34 \tab 2 \tab 27 \tab 11 \tab 26 \tab 56 \tab 3 \tab 9 \tab 93 \tab 40 \tab 8 \tab 112 \tab 5 \tab 50385 \tab 9 \tab 6 \tab 23 \tab 1 \tab 8 \tab 3 \tab 32 \tab 11 \tab 5 \tab 10 \tab 9 \tab 7 \tab 7 \tab 8 \tab 12 \tab 7 \tab 1 \tab 6 \tab 28 \tab 17 \tab 51 \tab 36 \tab 26 \tab 36 \tab 9 \tab 13 \tab 23 \tab 5 \tab 5 \tab 20 \tab 33 \tab 70 \tab 14 \tab 5 \tab 17 \tab 83 \tab 16 \tab 5 \tab 11 \tab 15 \tab 14 \tab 4 \tab 12 \tab 3 \tab 4 \tab 16 \tab 15 \tab 2 \tab 8 \tab 8 \tab 24 \tab 9 \tab 7 \tab 7 \tab 20 \tab 50 \tab 11 \tab 29 \tab 6 \tab 14 \tab 9 \tab 6 \tab 2 \tab 3 \tab 3 \tab 5 \tab 8 \tab 15 \tab 8 \tab 5 \tab 2 \tab 17 \tab 7 \tab 10 \tab 21 \tab 11 \tab 11 \tab 29 \tab 26 \tab 29 \tab 8 \tab 15 \tab 9 \tab 27 \tab 13 \tab 6 \tab 11 \tab 2 \tab 5 \tab 1 \tab 5 \tab 8 \tab 6 \tab 9 \tab 9 \tab 1 \tab 5 \tab 13 \tab 6 \tab 10 \tab 19 \tab 3 \tab 15 \tab 41 \tab 3 \tab 4 \tab 7 \tab 4 \tab 2 \tab 4 \tab 18 \tab 5 \tab 9 \tab 4 \tab 8 \tab 4 \tab 12 \tab 1 \tab 4 \tab 5 \tab 4 \tab 2 \tab 4 \tab 8 \tab 15 \tab 3 \tab 5 \tab 1 \tab 3 \tab 8 \tab 1 \tab 3 \tab 4 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
