% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_19.R
\name{padotsvp}
\alias{padotsvp}
\title{Fa. d.o.t. code - specific vocat. prep}
\source{
General Social Survey https://gss.norc.org
}
\description{
padotsvp
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 1 \tab 1.56 \tab 1.7 \tab 2 \tab 2.03 \tab 2.15 \tab 2.19 \tab 2.27 \tab 2.29 \tab 2.33 \tab 2.34 \tab 2.36 \tab 2.44 \tab 2.5 \tab 2.55 \tab 2.56 \tab 2.57 \tab 2.7 \tab 2.72 \tab 2.77 \tab 2.84 \tab 2.86 \tab 2.9 \tab 2.91 \tab 2.94 \tab 2.95 \tab 2.99 \tab 3.02 \tab 3.08 \tab 3.09 \tab 3.12 \tab 3.13 \tab 3.15 \tab 3.19 \tab 3.24 \tab 3.28 \tab 3.29 \tab 3.3 \tab 3.32 \tab 3.33 \tab 3.34 \tab 3.35 \tab 3.43 \tab 3.45 \tab 3.46 \tab 3.47 \tab 3.5 \tab 3.51 \tab 3.58 \tab 3.59 \tab 3.62 \tab 3.63 \tab 3.69 \tab 3.72 \tab 3.73 \tab 3.75 \tab 3.76 \tab 3.79 \tab 3.8 \tab 3.81 \tab 3.82 \tab 3.84 \tab 3.88 \tab 3.91 \tab 3.93 \tab 3.95 \tab 3.96 \tab 3.97 \tab 4 \tab 4.02 \tab 4.03 \tab 4.09 \tab 4.1 \tab 4.12 \tab 4.14 \tab 4.16 \tab 4.17 \tab 4.22 \tab 4.24 \tab 4.26 \tab 4.27 \tab 4.34 \tab 4.4 \tab 4.44 \tab 4.51 \tab 4.53 \tab 4.56 \tab 4.58 \tab 4.6 \tab 4.64 \tab 4.67 \tab 4.69 \tab 4.74 \tab 4.79 \tab 4.84 \tab 4.86 \tab 4.88 \tab 4.95 \tab 4.96 \tab 4.97 \tab 4.98 \tab 5 \tab 5.01 \tab 5.02 \tab 5.04 \tab 5.05 \tab 5.13 \tab 5.14 \tab 5.19 \tab 5.38 \tab 5.42 \tab 5.49 \tab 5.5 \tab 5.54 \tab 5.62 \tab 5.67 \tab 5.75 \tab 5.82 \tab 5.84 \tab 5.87 \tab 5.89 \tab 5.9 \tab 5.95 \tab 5.98 \tab 5.99 \tab 6.01 \tab 6.08 \tab 6.14 \tab 6.17 \tab 6.18 \tab 6.25 \tab 6.26 \tab 6.27 \tab 6.38 \tab 6.39 \tab 6.4 \tab 6.47 \tab 6.5 \tab 6.53 \tab 6.56 \tab 6.61 \tab 6.62 \tab 6.63 \tab 6.64 \tab 6.66 \tab 6.69 \tab 6.7 \tab 6.72 \tab 6.73 \tab 6.74 \tab 6.75 \tab 6.76 \tab 6.77 \tab 6.79 \tab 6.81 \tab 6.82 \tab 6.85 \tab 6.86 \tab 6.87 \tab 6.88 \tab 6.92 \tab 6.93 \tab 6.94 \tab 6.95 \tab 6.96 \tab 6.97 \tab 6.98 \tab 7 \tab 7.01 \tab 7.02 \tab 7.03 \tab 7.05 \tab 7.06 \tab 7.07 \tab 7.08 \tab 7.09 \tab 7.12 \tab 7.14 \tab 7.26 \tab 7.27 \tab 7.28 \tab 7.29 \tab 7.31 \tab 7.32 \tab 7.34 \tab 7.37 \tab 7.39 \tab 7.45 \tab 7.51 \tab 7.52 \tab 7.54 \tab 7.55 \tab 7.61 \tab 7.67 \tab 7.7 \tab 7.74 \tab 7.77 \tab 7.78 \tab 7.79 \tab 7.81 \tab 7.82 \tab 7.84 \tab 7.85 \tab 7.86 \tab 7.88 \tab 7.92 \tab 7.93 \tab 7.96 \tab 8 \tab iap \tab 2.14 \tab 2.39 \tab 2.54 \tab 2.64 \tab 2.74 \tab 2.75 \tab 2.76 \tab 2.92 \tab 3.14 \tab 3.21 \tab 3.25 \tab 3.27 \tab 3.38 \tab 3.4 \tab 3.74 \tab 3.78 \tab 3.86 \tab 3.87 \tab 4.06 \tab 4.29 \tab 4.45 \tab 4.46 \tab 4.49 \tab 4.57 \tab 4.71 \tab 4.72 \tab 4.77 \tab 4.82 \tab 4.9 \tab 4.94 \tab 5.06 \tab 5.16 \tab 5.48 \tab 5.58 \tab 5.61 \tab 5.66 \tab 5.93 \tab 6 \tab 6.11 \tab 6.19 \tab 6.2 \tab 6.44 \tab 6.54 \tab 6.8 \tab 6.83 \tab 6.91 \tab 7.21 \tab 7.22 \tab 7.23 \tab 7.46 \tab 7.57 \tab 7.59 \tab 7.64 \tab 2.09 \tab 2.2 \tab 2.49 \tab 2.93 \tab 3.03 \tab 3.37 \tab 4.76 \tab 4.93 \tab 5.32 \tab 5.51 \tab 5.59 \tab 5.7 \tab 6.33 \tab 6.37 \tab 6.48 \tab 7.04 \tab 7.62 \tab 2.69 \tab 2.82 \tab 2.98 \tab 3.05 \tab 3.22 \tab 3.7 \tab 3.77 \tab 4.36 \tab 5.17 \tab 5.27 \tab 5.74 \tab 7.1 \tab 7.58 \tab 7.73 \tab 2.32 \tab 3 \tab 3.18 \tab 3.89 \tab 4.39 \tab 5.88 \tab 6.84 \tab 7.24 \tab 7.72 \tab 1.85 \tab 3.61 \tab 7.95 \tab 2.51 \tab 3.36 \tab 3.48 \tab 3.57 \tab 5.79 \tab 7.33 \tab 3.39 \tab 5.63 \tab 6.6 \tab 6.9 \tab 7.16 \tab 7.83 \tab 5.71 \tab 7.68 \tab 7.91 \tab 2.85 \tab 2.11 \tab 4.35 \tab 5.81 \tab 5.85 \tab 5.94 \tab 5.3 \tab 6.42 \tab 3.31 \tab Total \cr
   1972 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 2 \tab 3 \tab 1 \tab 3 \tab 17 \tab 3 \tab 6 \tab 3 \tab 3 \tab 3 \tab 6 \tab 6 \tab 3 \tab 6 \tab 4 \tab 4 \tab 1 \tab 2 \tab 9 \tab 29 \tab 2 \tab 13 \tab 6 \tab 2 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab 18 \tab 3 \tab 2 \tab 1 \tab 11 \tab 40 \tab 2 \tab 2 \tab 2 \tab 6 \tab 6 \tab 1 \tab 1 \tab 3 \tab 3 \tab 6 \tab 2 \tab 4 \tab 7 \tab 2 \tab 2 \tab 7 \tab 22 \tab 25 \tab 1 \tab 2 \tab 1 \tab 2 \tab 7 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 9 \tab 1 \tab 4 \tab 2 \tab 3 \tab 1 \tab 5 \tab 2 \tab 5 \tab 1 \tab 1 \tab 9 \tab 1 \tab 4 \tab 4 \tab 1 \tab 3 \tab 2 \tab 3 \tab 1 \tab 3 \tab 5 \tab 2 \tab 12 \tab 10 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 11 \tab 1 \tab 1 \tab 6 \tab 4 \tab 1 \tab 1 \tab 5 \tab 4 \tab 2 \tab 5 \tab 1 \tab 7 \tab 1 \tab 1 \tab 9 \tab 1 \tab 2 \tab 8 \tab 1 \tab 3 \tab 6 \tab 2 \tab 7 \tab 1 \tab 4 \tab 79 \tab 308 \tab 2 \tab 3 \tab 17 \tab 1 \tab 2 \tab 12 \tab 1 \tab 6 \tab 1 \tab 9 \tab 3 \tab 2 \tab 11 \tab 4 \tab 4 \tab 9 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 21 \tab 5 \tab 15 \tab 3 \tab 8 \tab 2 \tab 4 \tab 3 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 12 \tab 2 \tab 2 \tab 16 \tab 9 \tab 1 \tab 27 \tab 2 \tab 1 \tab 1 \tab 2 \tab 9 \tab 1 \tab 3 \tab 2 \tab 35 \tab 11 \tab 3 \tab 1 \tab 2 \tab 3 \tab 7 \tab 5 \tab 6 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 12 \tab 10 \tab 7 \tab 276 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 5 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab 10 \tab - \tab 7 \tab 2 \tab 2 \tab 1 \tab 7 \tab 12 \tab 2 \tab 7 \tab 5 \tab 4 \tab 2 \tab 4 \tab 17 \tab 25 \tab - \tab 9 \tab 4 \tab - \tab - \tab 5 \tab 3 \tab 5 \tab 3 \tab 2 \tab 8 \tab - \tab - \tab - \tab 7 \tab 34 \tab - \tab 6 \tab 5 \tab 3 \tab 6 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 9 \tab 2 \tab 2 \tab 2 \tab 6 \tab 15 \tab 34 \tab - \tab 2 \tab 3 \tab - \tab 11 \tab - \tab 3 \tab 4 \tab 3 \tab - \tab 11 \tab - \tab 5 \tab 2 \tab 2 \tab 3 \tab 6 \tab 4 \tab 1 \tab 2 \tab - \tab 6 \tab 1 \tab - \tab - \tab 9 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 6 \tab - \tab 6 \tab 9 \tab - \tab 5 \tab 1 \tab - \tab - \tab 6 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab 2 \tab 5 \tab 1 \tab 2 \tab 9 \tab - \tab 6 \tab 2 \tab - \tab 7 \tab - \tab 1 \tab 7 \tab 4 \tab - \tab 6 \tab 3 \tab 2 \tab 2 \tab 1 \tab 82 \tab 256 \tab - \tab 4 \tab 10 \tab 2 \tab - \tab 13 \tab - \tab 6 \tab 1 \tab 5 \tab 2 \tab 3 \tab 11 \tab 4 \tab 4 \tab 19 \tab 1 \tab - \tab 6 \tab 1 \tab 2 \tab 22 \tab 4 \tab 16 \tab 4 \tab 6 \tab - \tab 2 \tab 7 \tab 7 \tab 1 \tab 2 \tab 3 \tab - \tab 6 \tab 11 \tab - \tab - \tab 10 \tab 15 \tab - \tab 20 \tab 4 \tab 5 \tab 2 \tab 6 \tab 9 \tab 2 \tab - \tab 2 \tab 39 \tab 9 \tab 2 \tab - \tab - \tab 3 \tab 7 \tab 5 \tab 5 \tab 2 \tab 5 \tab - \tab 3 \tab 3 \tab 9 \tab 12 \tab 2 \tab 191 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 42 \tab 8 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 2 \tab - \tab 6 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 17 \tab - \tab 5 \tab 1 \tab - \tab 2 \tab 3 \tab 7 \tab 1 \tab 3 \tab 8 \tab 4 \tab 1 \tab - \tab 10 \tab 16 \tab - \tab 3 \tab 14 \tab - \tab - \tab 2 \tab - \tab 1 \tab 4 \tab 8 \tab 9 \tab - \tab - \tab - \tab 7 \tab 37 \tab 2 \tab 1 \tab 1 \tab 7 \tab 12 \tab 2 \tab 1 \tab 4 \tab 4 \tab 5 \tab 1 \tab 3 \tab 2 \tab 3 \tab 1 \tab 16 \tab 3 \tab 22 \tab 1 \tab 1 \tab - \tab - \tab 6 \tab - \tab 1 \tab 7 \tab 6 \tab 1 \tab 11 \tab 2 \tab 7 \tab 5 \tab 2 \tab - \tab 4 \tab - \tab 4 \tab 2 \tab - \tab 7 \tab 2 \tab 2 \tab 4 \tab 8 \tab 4 \tab 1 \tab 2 \tab - \tab 3 \tab 8 \tab - \tab 8 \tab 12 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab 5 \tab 5 \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab 5 \tab 11 \tab 1 \tab - \tab 4 \tab - \tab 8 \tab - \tab 1 \tab 8 \tab 1 \tab - \tab 4 \tab 2 \tab 5 \tab 2 \tab 3 \tab 83 \tab 277 \tab - \tab 7 \tab 21 \tab 3 \tab - \tab 8 \tab 1 \tab 4 \tab - \tab 5 \tab 1 \tab 4 \tab 16 \tab 1 \tab 2 \tab 19 \tab - \tab 2 \tab 3 \tab 1 \tab 2 \tab 14 \tab 2 \tab 10 \tab 3 \tab 1 \tab - \tab 4 \tab 2 \tab 9 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 10 \tab - \tab 2 \tab 20 \tab 13 \tab 1 \tab 27 \tab - \tab 3 \tab - \tab 7 \tab 2 \tab - \tab 2 \tab 1 \tab 44 \tab 6 \tab 10 \tab - \tab 2 \tab 1 \tab 8 \tab 2 \tab 3 \tab 3 \tab 4 \tab 3 \tab 1 \tab 2 \tab 10 \tab 7 \tab 2 \tab 215 \tab - \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 3 \tab 1 \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1 \tab 3 \tab - \tab 3 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 17 \tab - \tab 4 \tab 2 \tab 3 \tab 1 \tab 3 \tab 4 \tab - \tab 11 \tab 15 \tab 3 \tab - \tab 1 \tab 22 \tab 31 \tab - \tab 2 \tab 7 \tab - \tab 2 \tab 1 \tab - \tab 5 \tab 4 \tab - \tab 12 \tab - \tab - \tab - \tab 4 \tab 15 \tab 2 \tab 3 \tab - \tab 7 \tab 13 \tab 2 \tab 1 \tab 1 \tab 3 \tab 8 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab 24 \tab 29 \tab 25 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab - \tab - \tab 9 \tab 2 \tab 1 \tab 10 \tab - \tab 4 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 3 \tab 3 \tab - \tab - \tab - \tab 1 \tab 11 \tab 21 \tab - \tab 1 \tab - \tab 1 \tab - \tab 12 \tab 5 \tab 4 \tab 10 \tab 1 \tab - \tab 8 \tab 8 \tab - \tab 3 \tab 7 \tab 1 \tab 3 \tab 1 \tab - \tab 9 \tab - \tab - \tab 5 \tab 4 \tab - \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 75 \tab 270 \tab - \tab 7 \tab 21 \tab - \tab 3 \tab 13 \tab 2 \tab 5 \tab - \tab 5 \tab 7 \tab 2 \tab 4 \tab 1 \tab 1 \tab 10 \tab 3 \tab 1 \tab 3 \tab - \tab 3 \tab 20 \tab 3 \tab 24 \tab 4 \tab 10 \tab 2 \tab 2 \tab 5 \tab 4 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 12 \tab 3 \tab 2 \tab 18 \tab 9 \tab 3 \tab 25 \tab 3 \tab 2 \tab 1 \tab 7 \tab 5 \tab - \tab 1 \tab 1 \tab 44 \tab 8 \tab 5 \tab - \tab 1 \tab - \tab 8 \tab 7 \tab 6 \tab 2 \tab 3 \tab 4 \tab 1 \tab 1 \tab 12 \tab 8 \tab 4 \tab 190 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 8 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1 \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 20 \tab - \tab 7 \tab 3 \tab 2 \tab 4 \tab 3 \tab 16 \tab 2 \tab 3 \tab 15 \tab 2 \tab - \tab - \tab 19 \tab 27 \tab 1 \tab 2 \tab 8 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab - \tab 17 \tab 3 \tab - \tab - \tab 4 \tab 17 \tab 1 \tab 4 \tab 1 \tab 5 \tab 8 \tab 1 \tab 3 \tab 2 \tab 3 \tab 4 \tab 1 \tab 14 \tab 2 \tab 1 \tab 1 \tab 19 \tab 28 \tab 23 \tab 1 \tab 3 \tab 1 \tab 2 \tab 5 \tab 1 \tab 1 \tab 5 \tab 5 \tab 3 \tab 8 \tab 1 \tab 8 \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 6 \tab - \tab 3 \tab 4 \tab - \tab 1 \tab 3 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 6 \tab - \tab 2 \tab - \tab - \tab - \tab 7 \tab - \tab 2 \tab 9 \tab - \tab - \tab 3 \tab 3 \tab 2 \tab 1 \tab 7 \tab 1 \tab 2 \tab 3 \tab - \tab 11 \tab - \tab 2 \tab 15 \tab 2 \tab - \tab 7 \tab 1 \tab 2 \tab 5 \tab 6 \tab 64 \tab 286 \tab - \tab 3 \tab 19 \tab 1 \tab - \tab 11 \tab 2 \tab 5 \tab - \tab 6 \tab - \tab 8 \tab 13 \tab 3 \tab 2 \tab 10 \tab 2 \tab 4 \tab 3 \tab - \tab - \tab 26 \tab 5 \tab 9 \tab 8 \tab 5 \tab - \tab 3 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 17 \tab - \tab 2 \tab 6 \tab 8 \tab 5 \tab 35 \tab 2 \tab 2 \tab 1 \tab 5 \tab 6 \tab 5 \tab 4 \tab - \tab 35 \tab 7 \tab 6 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 4 \tab 1 \tab 3 \tab 1 \tab - \tab 3 \tab 14 \tab 9 \tab 4 \tab 219 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 5 \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 21 \tab - \tab 9 \tab 1 \tab 2 \tab 1 \tab - \tab 5 \tab 2 \tab 6 \tab 10 \tab 4 \tab 2 \tab 2 \tab 20 \tab 25 \tab - \tab 2 \tab 10 \tab - \tab - \tab 2 \tab - \tab 3 \tab 5 \tab - \tab 20 \tab 1 \tab 1 \tab - \tab 3 \tab 39 \tab - \tab 4 \tab - \tab 7 \tab 13 \tab 1 \tab 2 \tab 5 \tab 3 \tab 3 \tab 2 \tab 5 \tab 5 \tab - \tab 1 \tab 12 \tab 21 \tab 44 \tab - \tab 6 \tab - \tab - \tab 5 \tab - \tab 1 \tab 4 \tab 2 \tab 2 \tab 11 \tab - \tab 3 \tab - \tab 1 \tab 3 \tab 3 \tab 1 \tab 4 \tab 4 \tab - \tab 6 \tab 1 \tab 5 \tab 1 \tab - \tab 2 \tab 2 \tab 3 \tab - \tab 4 \tab - \tab - \tab 4 \tab 24 \tab - \tab 1 \tab 2 \tab - \tab - \tab 7 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab 6 \tab 2 \tab 3 \tab 1 \tab 9 \tab 1 \tab 2 \tab 2 \tab 2 \tab 9 \tab - \tab 1 \tab 9 \tab - \tab 2 \tab 5 \tab 1 \tab 8 \tab 2 \tab 1 \tab 63 \tab 263 \tab 1 \tab 4 \tab 18 \tab - \tab 3 \tab 16 \tab - \tab 5 \tab - \tab 10 \tab 4 \tab 5 \tab 24 \tab 1 \tab 5 \tab 18 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 24 \tab 1 \tab 15 \tab 3 \tab 4 \tab 1 \tab 2 \tab 2 \tab 2 \tab 6 \tab 1 \tab 2 \tab - \tab 1 \tab 11 \tab 1 \tab - \tab 8 \tab 9 \tab 4 \tab 35 \tab 2 \tab 1 \tab - \tab 3 \tab 5 \tab - \tab - \tab 2 \tab 30 \tab 10 \tab 3 \tab - \tab 2 \tab 1 \tab 6 \tab 8 \tab 1 \tab 1 \tab 5 \tab 3 \tab - \tab 10 \tab 12 \tab 9 \tab 2 \tab 236 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 20 \tab 1 \tab 2 \tab - \tab 4 \tab 5 \tab 1 \tab 5 \tab 1 \tab 4 \tab 9 \tab 4 \tab - \tab 3 \tab 14 \tab 26 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 5 \tab 3 \tab - \tab 16 \tab 1 \tab - \tab 1 \tab 4 \tab 29 \tab 1 \tab 4 \tab - \tab 2 \tab 12 \tab - \tab 1 \tab 6 \tab - \tab 8 \tab 3 \tab 10 \tab 2 \tab 2 \tab 1 \tab 20 \tab 35 \tab 43 \tab 1 \tab 1 \tab 1 \tab 3 \tab 11 \tab - \tab 2 \tab 3 \tab 4 \tab 2 \tab 10 \tab 2 \tab 2 \tab 4 \tab - \tab 2 \tab 2 \tab 2 \tab 5 \tab 5 \tab - \tab 5 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab 7 \tab - \tab - \tab - \tab 1 \tab 4 \tab 17 \tab - \tab - \tab 3 \tab - \tab - \tab 5 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 8 \tab - \tab 9 \tab 5 \tab - \tab 9 \tab - \tab 5 \tab 7 \tab 5 \tab 1 \tab 3 \tab 1 \tab 8 \tab 1 \tab 4 \tab 73 \tab 241 \tab 1 \tab 3 \tab 12 \tab - \tab 1 \tab 12 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab 5 \tab 20 \tab 2 \tab 2 \tab 7 \tab - \tab 1 \tab 7 \tab - \tab 1 \tab 42 \tab 6 \tab 9 \tab 2 \tab 9 \tab - \tab 1 \tab 2 \tab 2 \tab 6 \tab 2 \tab 1 \tab 2 \tab 6 \tab 6 \tab 1 \tab 3 \tab 10 \tab 12 \tab 4 \tab 34 \tab 2 \tab 4 \tab 1 \tab 6 \tab 5 \tab 1 \tab - \tab 2 \tab 54 \tab 9 \tab 7 \tab - \tab 1 \tab 4 \tab 4 \tab 5 \tab 9 \tab - \tab 5 \tab 3 \tab 1 \tab 8 \tab 9 \tab 6 \tab - \tab 244 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 1 \tab - \tab 4 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 10 \tab 3 \tab 3 \tab 3 \tab - \tab 4 \tab 6 \tab 6 \tab 1 \tab 6 \tab 7 \tab 6 \tab - \tab 2 \tab 11 \tab 22 \tab 2 \tab 4 \tab 3 \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab 6 \tab - \tab 12 \tab - \tab - \tab - \tab 3 \tab 12 \tab 2 \tab 1 \tab 1 \tab 3 \tab 9 \tab 3 \tab 1 \tab - \tab 2 \tab 5 \tab - \tab 2 \tab 5 \tab - \tab - \tab 14 \tab 26 \tab 37 \tab - \tab 2 \tab 1 \tab 1 \tab 7 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 11 \tab 1 \tab 10 \tab 3 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 4 \tab - \tab 5 \tab - \tab - \tab 9 \tab 18 \tab - \tab 2 \tab 5 \tab 2 \tab - \tab 7 \tab 2 \tab - \tab 8 \tab 1 \tab - \tab 1 \tab 5 \tab 3 \tab 11 \tab 7 \tab 2 \tab 6 \tab 1 \tab 1 \tab 10 \tab 1 \tab 2 \tab 4 \tab 4 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 5 \tab 45 \tab 231 \tab 3 \tab 7 \tab 13 \tab - \tab 2 \tab 13 \tab - \tab 4 \tab 2 \tab 3 \tab - \tab 5 \tab 15 \tab 5 \tab 2 \tab 11 \tab 3 \tab - \tab 4 \tab 4 \tab 1 \tab 25 \tab 5 \tab 14 \tab 4 \tab 10 \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab 4 \tab 1 \tab - \tab 3 \tab 9 \tab 1 \tab 1 \tab 16 \tab 11 \tab 2 \tab 23 \tab 4 \tab 2 \tab 1 \tab 4 \tab 3 \tab 3 \tab 1 \tab 1 \tab 39 \tab 11 \tab 10 \tab - \tab - \tab 6 \tab 5 \tab 3 \tab 5 \tab 3 \tab 2 \tab - \tab 1 \tab 9 \tab 12 \tab 5 \tab 10 \tab 260 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 17 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 2 \tab 5 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 9 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 27 \tab 2 \tab 10 \tab - \tab 3 \tab 1 \tab 4 \tab 13 \tab 3 \tab 11 \tab 13 \tab 8 \tab - \tab 1 \tab 35 \tab 48 \tab 4 \tab 6 \tab 5 \tab - \tab 1 \tab 6 \tab - \tab 4 \tab 3 \tab - \tab 22 \tab 2 \tab 1 \tab - \tab 7 \tab 37 \tab 1 \tab 4 \tab - \tab 5 \tab 8 \tab 3 \tab 3 \tab 2 \tab 4 \tab 7 \tab - \tab 12 \tab 3 \tab 2 \tab 1 \tab 12 \tab 48 \tab 40 \tab - \tab 4 \tab - \tab - \tab 11 \tab - \tab 2 \tab 10 \tab 5 \tab 1 \tab 6 \tab 2 \tab 4 \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 4 \tab - \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 10 \tab 1 \tab 3 \tab - \tab 5 \tab - \tab 1 \tab 8 \tab 17 \tab - \tab 4 \tab 1 \tab 2 \tab 1 \tab 8 \tab 2 \tab 2 \tab 8 \tab 3 \tab - \tab - \tab 7 \tab 4 \tab 9 \tab 10 \tab 1 \tab 9 \tab 2 \tab 1 \tab 11 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 6 \tab 14 \tab 5 \tab - \tab 2 \tab 23 \tab 269 \tab 3 \tab 7 \tab 17 \tab - \tab 1 \tab 11 \tab 1 \tab 9 \tab 6 \tab 4 \tab 2 \tab 5 \tab 14 \tab 4 \tab 7 \tab 14 \tab 3 \tab - \tab - \tab 1 \tab - \tab 32 \tab 9 \tab 8 \tab 3 \tab 9 \tab - \tab 2 \tab 2 \tab 6 \tab 5 \tab - \tab - \tab 1 \tab 7 \tab 3 \tab 3 \tab 4 \tab 19 \tab 5 \tab 2 \tab 21 \tab 7 \tab 1 \tab 2 \tab 4 \tab 3 \tab 2 \tab - \tab 1 \tab 48 \tab 18 \tab 4 \tab 5 \tab - \tab 2 \tab 3 \tab 6 \tab 4 \tab 3 \tab 6 \tab 1 \tab 2 \tab 5 \tab 12 \tab 13 \tab 1 \tab 378 \tab - \tab 1 \tab - \tab - \tab 4 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 15 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 3 \tab 12 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 17 \tab - \tab 9 \tab 1 \tab 1 \tab 2 \tab 3 \tab 5 \tab 3 \tab 7 \tab 9 \tab 1 \tab - \tab 3 \tab 9 \tab 34 \tab - \tab 8 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 6 \tab - \tab 29 \tab - \tab - \tab - \tab 11 \tab 20 \tab 3 \tab 4 \tab - \tab 4 \tab 13 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 3 \tab 9 \tab - \tab 3 \tab 16 \tab 41 \tab 12 \tab - \tab 6 \tab - \tab 2 \tab 7 \tab - \tab - \tab 1 \tab 10 \tab 6 \tab 10 \tab - \tab 8 \tab 4 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 6 \tab 1 \tab 5 \tab 4 \tab 3 \tab 1 \tab - \tab 6 \tab - \tab 3 \tab - \tab 3 \tab - \tab 1 \tab 10 \tab 14 \tab - \tab 3 \tab 1 \tab - \tab - \tab 6 \tab 4 \tab 6 \tab 3 \tab 3 \tab 1 \tab 10 \tab 5 \tab 2 \tab 13 \tab 2 \tab 2 \tab 5 \tab 6 \tab 1 \tab 7 \tab 1 \tab - \tab 4 \tab 5 \tab 6 \tab 2 \tab 8 \tab 2 \tab 2 \tab - \tab 41 \tab 194 \tab 1 \tab 6 \tab 21 \tab 3 \tab - \tab 13 \tab - \tab 6 \tab 3 \tab 6 \tab 2 \tab - \tab 11 \tab 2 \tab 5 \tab 15 \tab 2 \tab - \tab 2 \tab 6 \tab 1 \tab 24 \tab 3 \tab 17 \tab 3 \tab 11 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 5 \tab 17 \tab 3 \tab 6 \tab 9 \tab 14 \tab 4 \tab 41 \tab 2 \tab 1 \tab 1 \tab 7 \tab 7 \tab 2 \tab 2 \tab 2 \tab 49 \tab 15 \tab 4 \tab 4 \tab - \tab 8 \tab 13 \tab 12 \tab 6 \tab 3 \tab 6 \tab 1 \tab 3 \tab 10 \tab 16 \tab 12 \tab 6 \tab 238 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 5 \tab - \tab 3 \tab 1 \tab 2 \tab 4 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 4 \tab 10 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 9 \tab - \tab 5 \tab - \tab 1 \tab 7 \tab 2 \tab - \tab - \tab 6 \tab 6 \tab 2 \tab - \tab 3 \tab 20 \tab 33 \tab - \tab 58 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 6 \tab - \tab 13 \tab - \tab 1 \tab - \tab 31 \tab 24 \tab - \tab 10 \tab - \tab 1 \tab 26 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 2 \tab - \tab 7 \tab 32 \tab - \tab - \tab - \tab 1 \tab 1 \tab 5 \tab - \tab - \tab 3 \tab 6 \tab 2 \tab 8 \tab 2 \tab 3 \tab 3 \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 6 \tab - \tab 4 \tab 2 \tab 3 \tab 1 \tab - \tab 7 \tab - \tab 4 \tab - \tab 5 \tab 1 \tab - \tab 6 \tab 9 \tab - \tab 7 \tab 4 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab 4 \tab 5 \tab 19 \tab 6 \tab - \tab 5 \tab 3 \tab - \tab 12 \tab - \tab 2 \tab 3 \tab 3 \tab 3 \tab 6 \tab - \tab 1 \tab 1 \tab 3 \tab 107 \tab 231 \tab - \tab 2 \tab 6 \tab - \tab - \tab 9 \tab 3 \tab 3 \tab - \tab 6 \tab 3 \tab 4 \tab 13 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab 6 \tab - \tab 2 \tab 23 \tab 5 \tab 7 \tab 2 \tab 4 \tab 1 \tab - \tab 2 \tab - \tab 6 \tab - \tab - \tab 1 \tab - \tab 27 \tab - \tab - \tab 17 \tab 5 \tab - \tab 53 \tab 2 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 21 \tab 11 \tab 3 \tab - \tab 2 \tab 8 \tab 5 \tab 3 \tab 8 \tab 1 \tab 4 \tab 1 \tab 2 \tab 6 \tab 7 \tab 5 \tab 3 \tab 249 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 3 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab 4 \tab 2 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 3 \tab - \tab 4 \tab 1 \tab - \tab 3 \tab 11 \tab 1 \tab 5 \tab 2 \tab 2 \tab 4 \tab 2 \tab 4 \tab - \tab 4 \tab 4 \tab 3 \tab 2 \tab 3 \tab 22 \tab 38 \tab 1 \tab 3 \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 14 \tab 1 \tab - \tab - \tab 4 \tab 23 \tab - \tab 2 \tab 1 \tab 6 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 5 \tab - \tab 11 \tab 6 \tab 1 \tab 2 \tab 4 \tab 30 \tab 14 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 12 \tab - \tab 4 \tab 8 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab - \tab 9 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 6 \tab - \tab 5 \tab - \tab - \tab 5 \tab 19 \tab - \tab 2 \tab 2 \tab - \tab - \tab 3 \tab - \tab 1 \tab 7 \tab 2 \tab 1 \tab 5 \tab 6 \tab 1 \tab 6 \tab 9 \tab 4 \tab 6 \tab 1 \tab - \tab 14 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 4 \tab 9 \tab - \tab 1 \tab 58 \tab 229 \tab 4 \tab 4 \tab 26 \tab 5 \tab 1 \tab 7 \tab 2 \tab 8 \tab 5 \tab 3 \tab 1 \tab 5 \tab 15 \tab 3 \tab 3 \tab 9 \tab 9 \tab 2 \tab 2 \tab 4 \tab 2 \tab 41 \tab 8 \tab 12 \tab 8 \tab 8 \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab - \tab 2 \tab - \tab 7 \tab 14 \tab 3 \tab 3 \tab 15 \tab 9 \tab 3 \tab 25 \tab 2 \tab 2 \tab 5 \tab 6 \tab 5 \tab 5 \tab 2 \tab 1 \tab 32 \tab 20 \tab 5 \tab 1 \tab 5 \tab 2 \tab 16 \tab 10 \tab 5 \tab 3 \tab 6 \tab 2 \tab 1 \tab 8 \tab 16 \tab 14 \tab 5 \tab 219 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 6 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 5 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 3 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 1 \tab - \tab 6 \tab - \tab - \tab 2 \tab 1 \tab - \tab 7 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab 10 \tab 3 \tab 5 \tab 2 \tab 4 \tab 6 \tab 6 \tab 4 \tab - \tab 5 \tab 3 \tab 11 \tab 1 \tab - \tab 9 \tab 32 \tab 1 \tab 6 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 18 \tab 1 \tab - \tab - \tab 3 \tab 23 \tab - \tab 2 \tab 2 \tab 3 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 3 \tab 6 \tab 2 \tab 1 \tab 1 \tab 1 \tab 56 \tab 10 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 4 \tab 2 \tab 12 \tab 1 \tab 5 \tab - \tab 4 \tab 2 \tab - \tab - \tab 2 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 4 \tab 1 \tab 4 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 21 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 9 \tab 1 \tab 3 \tab 8 \tab 4 \tab - \tab 5 \tab 2 \tab 3 \tab 6 \tab 2 \tab 3 \tab 5 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 56 \tab 218 \tab 2 \tab 1 \tab 8 \tab 2 \tab 2 \tab 12 \tab 2 \tab 3 \tab 8 \tab 3 \tab - \tab 9 \tab 9 \tab - \tab 3 \tab 17 \tab 4 \tab 3 \tab 1 \tab 7 \tab 2 \tab 34 \tab 2 \tab 26 \tab 1 \tab 11 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 4 \tab 13 \tab 4 \tab 2 \tab 12 \tab 4 \tab 4 \tab 33 \tab 1 \tab 2 \tab 3 \tab 7 \tab 4 \tab 1 \tab 2 \tab 1 \tab 33 \tab 3 \tab 10 \tab 2 \tab 2 \tab - \tab 6 \tab 3 \tab 8 \tab 3 \tab 6 \tab 2 \tab 1 \tab 8 \tab 11 \tab 8 \tab 9 \tab 258 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 18 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 5 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab 3 \tab 3 \tab 4 \tab 1 \tab 3 \tab 3 \tab 2 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab 6 \tab 11 \tab - \tab 5 \tab - \tab 2 \tab 4 \tab 6 \tab 11 \tab 1 \tab 6 \tab 9 \tab 3 \tab - \tab - \tab 42 \tab 41 \tab - \tab 8 \tab 7 \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 17 \tab 1 \tab - \tab - \tab 11 \tab 38 \tab 1 \tab 5 \tab 2 \tab 2 \tab 4 \tab 6 \tab 1 \tab 3 \tab 3 \tab 2 \tab 1 \tab 13 \tab 2 \tab 2 \tab 1 \tab 20 \tab 36 \tab 26 \tab - \tab 8 \tab - \tab 1 \tab 6 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 9 \tab 1 \tab 8 \tab 6 \tab 1 \tab 6 \tab 5 \tab 4 \tab 4 \tab - \tab - \tab 2 \tab - \tab 6 \tab 5 \tab - \tab 4 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 22 \tab - \tab 5 \tab 4 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 11 \tab 8 \tab 1 \tab 8 \tab 2 \tab - \tab 10 \tab - \tab 1 \tab 1 \tab 5 \tab - \tab 9 \tab 5 \tab 3 \tab 2 \tab 9 \tab 37 \tab 274 \tab 3 \tab 6 \tab 15 \tab 2 \tab - \tab 11 \tab 1 \tab 2 \tab 7 \tab 3 \tab 1 \tab 4 \tab 7 \tab 6 \tab 2 \tab 5 \tab 3 \tab - \tab 9 \tab 4 \tab 2 \tab 29 \tab 10 \tab 10 \tab 2 \tab 9 \tab - \tab 1 \tab 2 \tab 2 \tab 5 \tab 1 \tab 2 \tab - \tab 5 \tab 9 \tab 2 \tab 2 \tab 12 \tab 19 \tab 4 \tab 33 \tab 1 \tab - \tab 1 \tab 6 \tab 6 \tab 2 \tab - \tab 2 \tab 44 \tab 11 \tab 3 \tab 9 \tab 1 \tab - \tab 11 \tab 5 \tab 11 \tab 2 \tab 4 \tab 2 \tab 3 \tab 10 \tab 15 \tab 11 \tab 5 \tab 341 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 4 \tab 2 \tab - \tab 2 \tab 2 \tab 3 \tab 4 \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 6 \tab 8 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 7 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 5 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 9 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 12 \tab - \tab 11 \tab - \tab - \tab 5 \tab 9 \tab 1 \tab - \tab 8 \tab 8 \tab 8 \tab - \tab - \tab 18 \tab 35 \tab - \tab 7 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab 6 \tab - \tab 8 \tab 1 \tab - \tab - \tab 4 \tab 19 \tab 2 \tab 6 \tab - \tab - \tab 4 \tab 5 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 6 \tab 1 \tab 1 \tab 8 \tab 44 \tab 12 \tab - \tab 2 \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab 5 \tab 1 \tab 9 \tab - \tab 1 \tab 3 \tab 2 \tab 2 \tab 4 \tab - \tab 2 \tab 4 \tab - \tab 6 \tab 1 \tab 1 \tab 2 \tab 1 \tab 8 \tab - \tab 3 \tab - \tab 3 \tab - \tab 1 \tab 4 \tab 15 \tab - \tab 4 \tab 3 \tab 1 \tab - \tab 10 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab 3 \tab 2 \tab 1 \tab 6 \tab 3 \tab - \tab 8 \tab 2 \tab - \tab 3 \tab 3 \tab 5 \tab 7 \tab 2 \tab 5 \tab 1 \tab 4 \tab 59 \tab 209 \tab 1 \tab 6 \tab 18 \tab 2 \tab 8 \tab 7 \tab 2 \tab 7 \tab 3 \tab - \tab - \tab 6 \tab 6 \tab 2 \tab 3 \tab 12 \tab 2 \tab 3 \tab 3 \tab 2 \tab 1 \tab 27 \tab 6 \tab 25 \tab 1 \tab 9 \tab - \tab 1 \tab 2 \tab 2 \tab 7 \tab - \tab 4 \tab - \tab 5 \tab 7 \tab 2 \tab 3 \tab 9 \tab 11 \tab - \tab 39 \tab 1 \tab 3 \tab - \tab 4 \tab 5 \tab 1 \tab 3 \tab 3 \tab 29 \tab 9 \tab 5 \tab 1 \tab 5 \tab 2 \tab 10 \tab 3 \tab 7 \tab 2 \tab 8 \tab 1 \tab 5 \tab 4 \tab 19 \tab 10 \tab 7 \tab 242 \tab - \tab 5 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 5 \tab - \tab 3 \tab 8 \tab 15 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 9 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1481 \cr
   1989 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 21 \tab 4 \tab 6 \tab - \tab 1 \tab 2 \tab 4 \tab 6 \tab 1 \tab 6 \tab 11 \tab 3 \tab 1 \tab 6 \tab 29 \tab 45 \tab - \tab 5 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 4 \tab - \tab 13 \tab 2 \tab - \tab - \tab 2 \tab 16 \tab - \tab 5 \tab 1 \tab 6 \tab 4 \tab 3 \tab - \tab 4 \tab 1 \tab 4 \tab - \tab 11 \tab 5 \tab - \tab 2 \tab 8 \tab 29 \tab 21 \tab - \tab 8 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 5 \tab 1 \tab 11 \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 6 \tab 2 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 5 \tab 17 \tab - \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 12 \tab 2 \tab - \tab - \tab 6 \tab 1 \tab 3 \tab 6 \tab 1 \tab 6 \tab 5 \tab 4 \tab 5 \tab - \tab 1 \tab 4 \tab 3 \tab 1 \tab 4 \tab 5 \tab 2 \tab 6 \tab 7 \tab 70 \tab 206 \tab 1 \tab 6 \tab 24 \tab 3 \tab 1 \tab 5 \tab 1 \tab 6 \tab 1 \tab 5 \tab 3 \tab 4 \tab 19 \tab 2 \tab 5 \tab 11 \tab 2 \tab - \tab 5 \tab 5 \tab 1 \tab 32 \tab 12 \tab 16 \tab 6 \tab 3 \tab 1 \tab 3 \tab 4 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 6 \tab 12 \tab 2 \tab - \tab 11 \tab 10 \tab 2 \tab 52 \tab 2 \tab 2 \tab 2 \tab - \tab 4 \tab 4 \tab 3 \tab - \tab 24 \tab 11 \tab 6 \tab 5 \tab 3 \tab 5 \tab 9 \tab 7 \tab 7 \tab 4 \tab 10 \tab 3 \tab 2 \tab 6 \tab 13 \tab 12 \tab 6 \tab 256 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1537 \cr
   1990 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 11 \tab 1 \tab 7 \tab - \tab 1 \tab 2 \tab 1 \tab 5 \tab - \tab 3 \tab 11 \tab 3 \tab - \tab 3 \tab 26 \tab 37 \tab - \tab 3 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 18 \tab 1 \tab 1 \tab 1 \tab 2 \tab 17 \tab - \tab - \tab - \tab 6 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 7 \tab - \tab 7 \tab 1 \tab 3 \tab 1 \tab 8 \tab 39 \tab 9 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab 6 \tab - \tab 7 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 6 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab 2 \tab 21 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 3 \tab 2 \tab 2 \tab 5 \tab 1 \tab - \tab 4 \tab 3 \tab 2 \tab 3 \tab 4 \tab 1 \tab 3 \tab 3 \tab 1 \tab 8 \tab - \tab - \tab 2 \tab - \tab - \tab 6 \tab 3 \tab 3 \tab 2 \tab 4 \tab 67 \tab 165 \tab 2 \tab 11 \tab 21 \tab - \tab 3 \tab 11 \tab - \tab 6 \tab 6 \tab 1 \tab - \tab 4 \tab 12 \tab - \tab 2 \tab 13 \tab 4 \tab 2 \tab 2 \tab 5 \tab 1 \tab 32 \tab 5 \tab 12 \tab 3 \tab 8 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 10 \tab 4 \tab 2 \tab 12 \tab 9 \tab 4 \tab 38 \tab 3 \tab 3 \tab 2 \tab 2 \tab 4 \tab 2 \tab 1 \tab - \tab 28 \tab 7 \tab 11 \tab - \tab 1 \tab 3 \tab 15 \tab 5 \tab 9 \tab 6 \tab 10 \tab 1 \tab 1 \tab 1 \tab 9 \tab 8 \tab 4 \tab 248 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 10 \tab 34 \tab 5 \tab 44 \tab 20 \tab 5 \tab 40 \tab 4 \tab 13 \tab 38 \tab 14 \tab 28 \tab 18 \tab 13 \tab 32 \tab 47 \tab 11 \tab 21 \tab 12 \tab 13 \tab 15 \tab 18 \tab 5 \tab 37 \tab 261 \tab 18 \tab 106 \tab 20 \tab 31 \tab 54 \tab 66 \tab 110 \tab 20 \tab 102 \tab 147 \tab 73 \tab 10 \tab 33 \tab 332 \tab 544 \tab 12 \tab 140 \tab 84 \tab 4 \tab 13 \tab 37 \tab 12 \tab 37 \tab 59 \tab 11 \tab 264 \tab 17 \tab 6 \tab 3 \tab 118 \tab 440 \tab 17 \tab 63 \tab 16 \tab 73 \tab 144 \tab 35 \tab 23 \tab 43 \tab 35 \tab 70 \tab 19 \tab 113 \tab 64 \tab 22 \tab 24 \tab 202 \tab 534 \tab 397 \tab 4 \tab 55 \tab 10 \tab 16 \tab 98 \tab 2 \tab 17 \tab 53 \tab 70 \tab 29 \tab 164 \tab 14 \tab 85 \tab 56 \tab 24 \tab 27 \tab 42 \tab 17 \tab 44 \tab 50 \tab 2 \tab 88 \tab 23 \tab 33 \tab 33 \tab 23 \tab 65 \tab 17 \tab 55 \tab 2 \tab 43 \tab 22 \tab 9 \tab 99 \tab 272 \tab 1 \tab 46 \tab 32 \tab 16 \tab 4 \tab 108 \tab 25 \tab 38 \tab 99 \tab 31 \tab 6 \tab 52 \tab 72 \tab 37 \tab 98 \tab 112 \tab 21 \tab 88 \tab 45 \tab 12 \tab 153 \tab 6 \tab 20 \tab 82 \tab 46 \tab 27 \tab 77 \tab 55 \tab 67 \tab 30 \tab 58 \tab 1082 \tab 4127 \tab 24 \tab 87 \tab 287 \tab 24 \tab 27 \tab 184 \tab 18 \tab 86 \tab 45 \tab 77 \tab 29 \tab 75 \tab 220 \tab 40 \tab 54 \tab 203 \tab 43 \tab 21 \tab 60 \tab 41 \tab 24 \tab 468 \tab 91 \tab 245 \tab 60 \tab 125 \tab 10 \tab 31 \tab 40 \tab 53 \tab 55 \tab 20 \tab 25 \tab 8 \tab 63 \tab 200 \tab 31 \tab 34 \tab 220 \tab 172 \tab 43 \tab 561 \tab 40 \tab 39 \tab 23 \tab 76 \tab 82 \tab 31 \tab 24 \tab 22 \tab 628 \tab 176 \tab 97 \tab 28 \tab 27 \tab 50 \tab 135 \tab 92 \tab 104 \tab 40 \tab 90 \tab 29 \tab 29 \tab 97 \tab 208 \tab 159 \tab 77 \tab 50385 \tab 1 \tab 16 \tab 6 \tab 7 \tab 27 \tab 15 \tab 11 \tab 7 \tab 7 \tab 8 \tab 9 \tab 8 \tab 4 \tab 6 \tab 14 \tab 7 \tab 18 \tab 11 \tab 15 \tab 22 \tab 13 \tab 15 \tab 11 \tab 19 \tab 28 \tab 21 \tab 26 \tab 6 \tab 36 \tab 17 \tab 5 \tab 12 \tab 7 \tab 4 \tab 5 \tab 29 \tab 17 \tab 25 \tab 5 \tab 18 \tab 11 \tab 33 \tab 11 \tab 14 \tab 98 \tab 109 \tab 16 \tab 3 \tab 10 \tab 8 \tab 6 \tab 4 \tab 40 \tab 8 \tab 4 \tab 1 \tab 12 \tab 2 \tab 7 \tab 20 \tab 4 \tab 7 \tab 3 \tab 10 \tab 17 \tab 9 \tab 11 \tab 25 \tab 13 \tab 23 \tab 5 \tab 2 \tab 6 \tab 9 \tab 2 \tab 7 \tab 3 \tab 4 \tab 1 \tab 6 \tab 9 \tab 17 \tab 13 \tab 5 \tab 10 \tab 4 \tab 3 \tab 19 \tab 4 \tab 15 \tab 17 \tab 7 \tab 4 \tab 9 \tab 3 \tab 12 \tab 4 \tab 5 \tab 3 \tab 4 \tab 4 \tab 9 \tab 2 \tab 4 \tab 9 \tab 5 \tab 8 \tab 15 \tab 2 \tab 1 \tab 1 \tab 4 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
