% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_10.R
\name{paeduc}
\alias{paeduc}
\title{Highest year school completed, father}
\source{
General Social Survey https://gss.norc.org
}
\description{
paeduc
}
\details{
Question 16. What is the highest grade in elementary school or high school that your father finished and got credit for?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab iap \tab no answer \tab no formal schooling \tab Total \cr
   1972 \tab 6 \tab 58 \tab 44 \tab 195 \tab 24 \tab 40 \tab 9 \tab 59 \tab 5 \tab 17 \tab 3 \tab 12 \tab 12 \tab 51 \tab 57 \tab 43 \tab 68 \tab 49 \tab 231 \tab 34 \tab 297 \tab 235 \tab 10 \tab 54 \tab 1613 \cr
   1973 \tab 8 \tab 41 \tab 44 \tab 223 \tab 14 \tab 48 \tab 11 \tab 65 \tab 6 \tab 7 \tab 5 \tab 14 \tab 11 \tab 52 \tab 46 \tab 39 \tab 89 \tab 45 \tab 261 \tab 27 \tab 216 \tab 151 \tab 13 \tab 68 \tab 1504 \cr
   1974 \tab 8 \tab 55 \tab 37 \tab 234 \tab 21 \tab 48 \tab 11 \tab 50 \tab 4 \tab 15 \tab 1 \tab 10 \tab 10 \tab 49 \tab 46 \tab 43 \tab 81 \tab 47 \tab 213 \tab 43 \tab 223 \tab 182 \tab 1 \tab 52 \tab 1484 \cr
   1975 \tab 3 \tab 55 \tab 39 \tab 240 \tab 31 \tab 44 \tab 11 \tab 53 \tab 12 \tab 11 \tab 1 \tab 13 \tab 11 \tab 49 \tab 50 \tab 40 \tab 92 \tab 55 \tab 231 \tab 37 \tab 204 \tab 151 \tab 2 \tab 55 \tab 1490 \cr
   1976 \tab 7 \tab 45 \tab 41 \tab 243 \tab 21 \tab 40 \tab 8 \tab 48 \tab 9 \tab 15 \tab 4 \tab 12 \tab 15 \tab 44 \tab 47 \tab 39 \tab 91 \tab 50 \tab 223 \tab 37 \tab 217 \tab 186 \tab 9 \tab 48 \tab 1499 \cr
   1977 \tab 2 \tab 62 \tab 38 \tab 219 \tab 20 \tab 38 \tab 13 \tab 46 \tab 5 \tab 13 \tab 7 \tab 15 \tab 13 \tab 45 \tab 64 \tab 37 \tab 80 \tab 49 \tab 228 \tab 35 \tab 267 \tab 193 \tab 2 \tab 39 \tab 1530 \cr
   1978 \tab - \tab 58 \tab 29 \tab 256 \tab 29 \tab 51 \tab 8 \tab 54 \tab 16 \tab 12 \tab 6 \tab 11 \tab 10 \tab 36 \tab 39 \tab 35 \tab 88 \tab 53 \tab 213 \tab 44 \tab 235 \tab 205 \tab 3 \tab 41 \tab 1532 \cr
   1980 \tab 3 \tab 36 \tab 36 \tab 248 \tab 29 \tab 42 \tab 13 \tab 65 \tab 13 \tab 18 \tab 9 \tab 10 \tab 17 \tab 29 \tab 51 \tab 36 \tab 86 \tab 36 \tab 201 \tab 38 \tab 178 \tab 230 \tab 2 \tab 42 \tab 1468 \cr
   1982 \tab 3 \tab 66 \tab 36 \tab 309 \tab 29 \tab 55 \tab 14 \tab 72 \tab 10 \tab 20 \tab 6 \tab 7 \tab 17 \tab 42 \tab 46 \tab 40 \tab 74 \tab 49 \tab 226 \tab 46 \tab 323 \tab 296 \tab 9 \tab 65 \tab 1860 \cr
   1983 \tab 3 \tab 53 \tab 39 \tab 302 \tab 30 \tab 57 \tab 5 \tab 102 \tab 9 \tab 22 \tab 7 \tab 12 \tab 18 \tab 28 \tab 32 \tab 47 \tab 68 \tab 39 \tab 202 \tab 52 \tab 224 \tab 207 \tab - \tab 41 \tab 1599 \cr
   1984 \tab 1 \tab 57 \tab 43 \tab 278 \tab 28 \tab 51 \tab 9 \tab 79 \tab 9 \tab 15 \tab 6 \tab 4 \tab 16 \tab 32 \tab 41 \tab 28 \tab 66 \tab 29 \tab 207 \tab 41 \tab 201 \tab 202 \tab 3 \tab 27 \tab 1473 \cr
   1985 \tab 3 \tab 67 \tab 45 \tab 288 \tab 29 \tab 64 \tab 13 \tab 88 \tab 7 \tab 26 \tab 8 \tab 14 \tab 25 \tab 29 \tab 36 \tab 30 \tab 77 \tab 39 \tab 194 \tab 40 \tab 187 \tab 194 \tab - \tab 31 \tab 1534 \cr
   1986 \tab 4 \tab 52 \tab 44 \tab 297 \tab 22 \tab 54 \tab 13 \tab 85 \tab 6 \tab 13 \tab 6 \tab 5 \tab 12 \tab 34 \tab 35 \tab 25 \tab 74 \tab 30 \tab 178 \tab 30 \tab 193 \tab 222 \tab 4 \tab 32 \tab 1470 \cr
   1987 \tab 5 \tab 56 \tab 45 \tab 318 \tab 29 \tab 63 \tab 15 \tab 98 \tab 15 \tab 19 \tab 16 \tab 8 \tab 28 \tab 48 \tab 44 \tab 45 \tab 83 \tab 39 \tab 223 \tab 52 \tab 234 \tab 291 \tab 12 \tab 33 \tab 1819 \cr
   1988 \tab 1 \tab 49 \tab 30 \tab 291 \tab 33 \tab 43 \tab 11 \tab 94 \tab 10 \tab 28 \tab 4 \tab 11 \tab 27 \tab 41 \tab 24 \tab 33 \tab 59 \tab 26 \tab 183 \tab 39 \tab 208 \tab 189 \tab 19 \tab 28 \tab 1481 \cr
   1989 \tab 4 \tab 57 \tab 33 \tab 302 \tab 33 \tab 58 \tab 16 \tab 96 \tab 10 \tab 30 \tab 7 \tab 14 \tab 24 \tab 43 \tab 28 \tab 35 \tab 59 \tab 36 \tab 180 \tab 40 \tab 172 \tab 197 \tab 1 \tab 62 \tab 1537 \cr
   1990 \tab 2 \tab 42 \tab 20 \tab 300 \tab 35 \tab 54 \tab 17 \tab 85 \tab 9 \tab 21 \tab 6 \tab 11 \tab 20 \tab 33 \tab 29 \tab 17 \tab 42 \tab 25 \tab 159 \tab 33 \tab 165 \tab 190 \tab 35 \tab 22 \tab 1372 \cr
   1991 \tab - \tab 49 \tab 38 \tab 300 \tab 28 \tab 77 \tab 12 \tab 103 \tab 12 \tab 24 \tab 13 \tab 7 \tab 22 \tab 31 \tab 22 \tab 22 \tab 61 \tab 27 \tab 165 \tab 39 \tab 211 \tab 205 \tab 32 \tab 17 \tab 1517 \cr
   1993 \tab 2 \tab 62 \tab 34 \tab 317 \tab 37 \tab 79 \tab 20 \tab 123 \tab 12 \tab 36 \tab 9 \tab 9 \tab 31 \tab 39 \tab 22 \tab 23 \tab 63 \tab 34 \tab 187 \tab 34 \tab 185 \tab 197 \tab 29 \tab 22 \tab 1606 \cr
   1994 \tab 5 \tab 112 \tab 53 \tab 637 \tab 57 \tab 161 \tab 22 \tab 216 \tab 19 \tab 62 \tab 21 \tab 17 \tab 59 \tab 47 \tab 35 \tab 41 \tab 109 \tab 60 \tab 307 \tab 72 \tab 330 \tab 492 \tab 15 \tab 43 \tab 2992 \cr
   1996 \tab 2 \tab 101 \tab 59 \tab 631 \tab 72 \tab 148 \tab 40 \tab 229 \tab 22 \tab 60 \tab 15 \tab 6 \tab 69 \tab 40 \tab 40 \tab 43 \tab 98 \tab 43 \tab 254 \tab 63 \tab 324 \tab 502 \tab 12 \tab 31 \tab 2904 \cr
   1998 \tab 4 \tab 76 \tab 56 \tab 607 \tab 58 \tab 145 \tab 26 \tab 256 \tab 31 \tab 57 \tab 16 \tab 15 \tab 58 \tab 45 \tab 40 \tab 42 \tab 96 \tab 53 \tab 274 \tab 58 \tab 275 \tab 503 \tab 13 \tab 28 \tab 2832 \cr
   2000 \tab 4 \tab 95 \tab 59 \tab 627 \tab 65 \tab 129 \tab 26 \tab 212 \tab 13 \tab 63 \tab 15 \tab 19 \tab 55 \tab 44 \tab 36 \tab 43 \tab 95 \tab 35 \tab 227 \tab 52 \tab 326 \tab 518 \tab 27 \tab 32 \tab 2817 \cr
   2002 \tab 6 \tab 123 \tab 59 \tab 661 \tab 62 \tab 158 \tab 27 \tab 210 \tab 18 \tab 61 \tab 18 \tab 19 \tab 50 \tab 38 \tab 28 \tab 34 \tab 87 \tab 32 \tab 234 \tab 75 \tab 185 \tab 526 \tab 19 \tab 35 \tab 2765 \cr
   2004 \tab 3 \tab 81 \tab 65 \tab 692 \tab 72 \tab 165 \tab 25 \tab 256 \tab 29 \tab 69 \tab 28 \tab 5 \tab 60 \tab 36 \tab 27 \tab 29 \tab 84 \tab 48 \tab 204 \tab 79 \tab 182 \tab 514 \tab 26 \tab 33 \tab 2812 \cr
   2006 \tab 11 \tab 87 \tab 56 \tab 684 \tab 83 \tab 168 \tab 35 \tab 253 \tab 28 \tab 75 \tab 25 \tab 24 \tab 76 \tab 65 \tab 36 \tab 40 \tab 103 \tab 48 \tab 220 \tab 70 \tab 185 \tab 2049 \tab 18 \tab 71 \tab 4510 \cr
   2008 \tab 1 \tab 57 \tab 48 \tab 502 \tab 56 \tab 99 \tab 22 \tab 187 \tab 15 \tab 44 \tab 16 \tab 9 \tab 38 \tab 37 \tab 22 \tab 26 \tab 70 \tab 22 \tab 138 \tab 41 \tab 177 \tab 354 \tab 5 \tab 37 \tab 2023 \cr
   2010 \tab 5 \tab 55 \tab 48 \tab 536 \tab 54 \tab 95 \tab 19 \tab 177 \tab 13 \tab 59 \tab 15 \tab 11 \tab 40 \tab 40 \tab 19 \tab 29 \tab 58 \tab 30 \tab 133 \tab 47 \tab 150 \tab 370 \tab 8 \tab 33 \tab 2044 \cr
   2012 \tab 6 \tab 60 \tab 44 \tab 475 \tab 51 \tab 95 \tab 15 \tab 190 \tab 17 \tab 44 \tab 10 \tab 11 \tab 59 \tab 30 \tab 22 \tab 28 \tab 51 \tab 20 \tab 122 \tab 54 \tab 193 \tab 338 \tab 3 \tab 36 \tab 1974 \cr
   2014 \tab 5 \tab 81 \tab 46 \tab 660 \tab 57 \tab 125 \tab 33 \tab 271 \tab 24 \tab 62 \tab 19 \tab 23 \tab 56 \tab 31 \tab 24 \tab 25 \tab 81 \tab 26 \tab 144 \tab 51 \tab 189 \tab 459 \tab 5 \tab 41 \tab 2538 \cr
   2016 \tab 8 \tab 74 \tab 65 \tab 774 \tab 72 \tab 147 \tab 35 \tab 274 \tab 28 \tab 67 \tab 16 \tab 11 \tab 59 \tab 40 \tab 26 \tab 24 \tab 80 \tab 32 \tab 157 \tab 60 \tab 221 \tab 550 \tab 4 \tab 43 \tab 2867 \cr
   2018 \tab 1 \tab 54 \tab 50 \tab 596 \tab 55 \tab 136 \tab 29 \tab 238 \tab 19 \tab 65 \tab 13 \tab 12 \tab 48 \tab 33 \tab 15 \tab 25 \tab 64 \tab 33 \tab 112 \tab 47 \tab 194 \tab 464 \tab 3 \tab 42 \tab 2348 \cr
   2021 \tab 2 \tab 102 \tab 69 \tab 1157 \tab 105 \tab 231 \tab 55 \tab 538 \tab 52 \tab 118 \tab 31 \tab 13 \tab 112 \tab 34 \tab 18 \tab 40 \tab 82 \tab 41 \tab 177 \tab 71 \tab 372 \tab 563 \tab 7 \tab 42 \tab 4032 \cr
   2022 \tab 5 \tab 91 \tab 62 \tab 1007 \tab 60 \tab 186 \tab 36 \tab 413 \tab 29 \tab 119 \tab 27 \tab 10 \tab 104 \tab 34 \tab 27 \tab 26 \tab 71 \tab 31 \tab 152 \tab 65 \tab - \tab 942 \tab - \tab 47 \tab 3544 \cr
   Total \tab 133 \tab 2269 \tab 1554 \tab 15406 \tab 1501 \tab 3194 \tab 674 \tab 5385 \tab 536 \tab 1387 \tab 409 \tab 404 \tab 1282 \tab 1349 \tab 1174 \tab 1152 \tab 2630 \tab 1311 \tab 6760 \tab 1646 \tab 7443 \tab 13067 \tab 351 \tab 1373 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} no formal schooling
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
