% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_552.R
\name{paidsex}
\alias{paidsex}
\title{R had sex for pay last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
paidsex
}
\details{
Question 1543. If you had other partners, please indicate all categories that apply to them. d. Person you paid or paid you for sex
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab don't know \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1243 \tab 166 \tab 68 \tab 4 \tab - \tab 1481 \cr
   1989 \tab 1303 \tab 156 \tab 74 \tab 4 \tab - \tab 1537 \cr
   1990 \tab 1155 \tab 142 \tab 71 \tab 4 \tab - \tab 1372 \cr
   1991 \tab 1293 \tab 150 \tab 67 \tab 7 \tab - \tab 1517 \cr
   1993 \tab 1359 \tab 156 \tab 87 \tab 4 \tab - \tab 1606 \cr
   1994 \tab 2489 \tab 328 \tab 169 \tab 6 \tab - \tab 2992 \cr
   1996 \tab 2401 \tab 378 \tab 119 \tab 6 \tab - \tab 2904 \cr
   1998 \tab 2383 \tab 283 \tab 154 \tab 12 \tab - \tab 2832 \cr
   2000 \tab 2335 \tab 299 \tab 177 \tab 6 \tab - \tab 2817 \cr
   2002 \tab 2317 \tab 395 \tab 44 \tab 9 \tab - \tab 2765 \cr
   2004 \tab 2429 \tab 367 \tab 6 \tab 10 \tab - \tab 2812 \cr
   2006 \tab 4087 \tab 395 \tab 17 \tab 11 \tab - \tab 4510 \cr
   2008 \tab 1688 \tab 298 \tab 24 \tab 13 \tab - \tab 2023 \cr
   2010 \tab 1705 \tab 296 \tab 30 \tab 13 \tab - \tab 2044 \cr
   2012 \tab 1669 \tab 259 \tab 35 \tab 11 \tab - \tab 1974 \cr
   2014 \tab 2144 \tab 325 \tab 52 \tab 17 \tab - \tab 2538 \cr
   2016 \tab 2530 \tab 306 \tab 20 \tab 11 \tab - \tab 2867 \cr
   2018 \tab 2130 \tab 193 \tab 20 \tab 3 \tab 2 \tab 2348 \cr
   2021 \tab 3747 \tab 243 \tab 20 \tab 11 \tab 11 \tab 4032 \cr
   2022 \tab 3316 \tab 152 \tab 41 \tab 14 \tab 21 \tab 3544 \cr
   Total \tab 65598 \tab 5287 \tab 1295 \tab 176 \tab 34 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
