% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{paind10}
\alias{paind10}
\title{paind10}
\source{
General Social Survey https://gss.norc.org
}
\description{
Father's industry code (naics 2007)
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab accounting, tax preparation, bookkeeping, and payroll services \tab administration of environmental quality and housing programs \tab administration of human resource programs \tab aerospace product and parts manufacturing \tab agricultural chemical manufacturing \tab air transportation \tab aircraft and parts manufacturing \tab alcoholic beverages, merchant wholesalers \tab animal food, grain, and oilseed milling \tab animal production \tab animal slaughtering and processing \tab apparel, fabrics, and notions, merchant wholesalers \tab architectural, engineering, and related services \tab auto parts, accessories, and tire stores \tab automobile dealers \tab automotive repair and maintenance \tab bakeries, except retail \tab banking and related activities \tab barber shops \tab beverage manufacturing \tab building material and supplies dealers \tab bus service and urban transit \tab carpet and rug mills \tab clothing stores \tab coal mining \tab colleges and universities, including junior colleges \tab commercial and industrial machinery and equipment repair and maintenance \tab commercial and service industry machinery manufacturing \tab communications, and audio and video equipment manufacturing \tab computer and peripheral equipment manufacturing \tab construction \tab construction, and mining and oil and gas field machinery manufacturing \tab crop production \tab cut and sew apparel manufacturing \tab cutlery and hand tool manufacturing \tab dairy product manufacturing \tab department stores and discount stores \tab don't know \tab drinking places, alcoholic beverages \tab drugs, sundries, and chemical and allied products, merchant wholesalers \tab drycleaning and laundry services \tab electric power generation, transmission, and distribution \tab electrical lighting and electrical equipment manufacturing, and other electrical component manufacturing, n.e.c. \tab electronic and precision equipment repair and maintenance \tab elementary and secondary schools \tab engines, turbines, and power transmission equipment manufacturing \tab executive offices and legislative bodies \tab fabric mills, except knitting mills \tab farm product raw materials, merchant wholesalers \tab fiber, yarn, and thread mills \tab fishing, hunting, and trapping \tab footwear and leather goods repair \tab footwear manufacturing \tab forestry, except logging \tab foundries \tab fuel dealers \tab funeral homes, and cemeteries and crematories \tab furniture and home furnishings stores \tab furniture and related product manufacturing \tab gasoline stations \tab glass and glass product manufacturing \tab groceries and related products, merchant wholesalers \tab grocery stores \tab hardware stores \tab hardware, plumbing and heating equipment, and supplies, merchant wholesalers \tab health and personal care, except drug, stores \tab hospitals \tab household appliance manufacturing \tab household appliance stores \tab iap \tab independent artists, performing arts, spectator sports, and related industries \tab individual and family services \tab industrial and miscellaneous chemicals \tab insurance carriers and related activities \tab investigation and security services \tab iron and steel mills and steel product manufacturing \tab jewelry, luggage, and leather goods stores \tab justice, public order, and safety activities \tab knitting fabric mills, and apparel knitting mills \tab labor unions \tab landscaping services \tab leather tanning and finishing and other allied products manufacturing \tab legal services \tab logging \tab lumber and other construction materials, merchant wholesalers \tab machinery manufacturing, n.e.c. \tab machinery, equipment, and supplies, merchant wholesalers \tab management, scientific, and technical consulting services \tab metal ore mining \tab metals and minerals, except petroleum, merchant wholesalers \tab metalworking machinery manufacturing \tab miscellaneous durable goods, merchant wholesalers \tab miscellaneous fabricated metal products manufacturing \tab miscellaneous general merchandise stores \tab miscellaneous manufacturing, n.e.c. \tab miscellaneous nonmetallic mineral product manufacturing \tab miscellaneous paper and pulp products \tab miscellaneous retail stores \tab miscellaneous wood products \tab motor vehicles and motor vehicle equipment manufacturing \tab museums, art galleries, historical sites, and similar institutions \tab national security and international affairs \tab natural gas distribution \tab navigational, measuring, electromedical, and control instruments manufacturing \tab newspaper publishers \tab no answer \tab non-depository credit and related activities \tab nonferrous metal (except aluminum) production and processing \tab nonmetallic mineral mining and quarrying \tab not specified food industries \tab not specified machinery manufacturing \tab not specified manufacturing industries \tab not specified metal industries \tab not specified retail trade \tab not specified type of mining \tab office supplies and stationery stores \tab offices of chiropractors \tab offices of dentists \tab offices of physicians \tab ordnance \tab other amusement, gambling, and recreation industries \tab other consumer goods rental \tab other direct selling establishments \tab other general government and support \tab other personal services \tab other professional, scientific, and technical services \tab paint, coating, and adhesive manufacturing \tab paper and paper products, merchant wholesalers \tab paperboard containers and boxes \tab periodical, book, and directory publishers \tab personal and household goods repair and maintenance \tab petroleum and petroleum products, merchant wholesalers \tab petroleum refining \tab pharmaceutical and medicine manufacturing \tab pharmacies and drug stores \tab plastics product manufacturing \tab postal service \tab pottery, ceramics, and plumbing fixture manufacturing \tab printing and related support activities \tab private households \tab pulp, paper, and paperboard mills \tab radio and television broadcasting and cable subscription programming \tab rail transportation \tab railroad rolling stock manufacturing \tab real estate \tab recyclable material, merchant wholesalers \tab religious organizations \tab restaurants and other food services \tab retail bakeries \tab rubber product, except tire, manufacturing \tab sawmills and wood preservation \tab scientific research and development services \tab securities, commodities, funds, trusts, and other financial investments \tab services incidental to transportation \tab services to buildings and dwellings \tab sewage treatment facilities \tab ship and boat building \tab soap, cleaning compound, and cosmetics manufacturing \tab specialized design services \tab specialty food stores \tab structural clay product manufacturing \tab structural metals, and boiler, tank, and shipping container manufacturing \tab taxi and limousine service \tab textile and fabric finishing and coating mills \tab textile product mills, except carpet and rug \tab tire manufacturing \tab tobacco manufacturing \tab traveler accommodation \tab truck transportation \tab u.s. air force \tab u.s. army \tab u.s. navy \tab uncodeable \tab warehousing and storage \tab waste management and remediation services \tab water transportation \tab water, steam, air-conditioning, and irrigation systems \tab wholesale electronic markets, agents and brokers \tab wired telecommunications carriers \tab active duty military personnel, all other \tab administration of economic programs and space research \tab advertising and related services \tab agricultural implement manufacturing \tab aluminum production and processing \tab apparel accessories and other apparel manufacturing \tab beer, wine, and liquor stores \tab business support services \tab cement, concrete, lime, and gypsum product manufacturing \tab civic, social, advocacy organizations, and grantmaking and giving services \tab electric and gas, and other combinations \tab electrical and electronic goods, merchant wholesalers \tab electronic component and product manufacturing, n.e.c. \tab fruit and vegetable preserving and specialty food manufacturing \tab lawn and garden equipment and supplies stores \tab machine shops; turned product; screw, nut, and bolt manufacturing \tab medical equipment and supplies manufacturing \tab miscellaneous nondurable goods, merchant wholesalers \tab motor vehicles, parts and supplies, merchant wholesalers \tab offices of optometrists \tab oil and gas extraction \tab other schools and instruction, and educational support services \tab radio, tv, and computer stores \tab residential care facilities, without nursing \tab resin, synthetic rubber and fibers, and filaments manufacturing \tab retail florists \tab savings institutions, including credit unions \tab seafood and other miscellaneous foods, n.e.c. \tab sewing, needlework, and piece goods stores \tab shoe stores \tab sporting and athletic goods, and doll, toy and game manufacturing \tab sporting goods, camera, and hobby and toy stores \tab sugar and confectionery products \tab support activities for agriculture and forestry \tab support activities for mining \tab veneer, plywood, and engineered wood products \tab veterinary services \tab bowling centers \tab business, professional, political, and similar organizations \tab farm supplies, merchant wholesalers \tab mail order houses \tab motion pictures and video industries \tab not specified utilities \tab not specified wholesale trade \tab other information services \tab pipeline transportation \tab public finance activities \tab u.s. armed forces/n.s. \tab u.s. marine corps \tab furniture and home furnishing, merchant wholesalers \tab metal forgings and stampings \tab other telecommunications services \tab prefabricated wood buildings and mobile homes \tab scenic and sightseeing transportation \tab used merchandise stores \tab vending machine operators \tab 1976 \tab automotive equipment rental and leasing \tab beauty salons \tab book stores and news dealers \tab coating, engraving, heat treating and allied activities \tab couriers and messengers \tab management of companies and enterprises \tab nursing care facilities \tab other transportation equipment manufacturing \tab professional and commercial equipment and supplies, merchant wholesalers \tab u.s. coast guard \tab business, technical, and trade schools and training \tab computer systems design and related services \tab employment services \tab commercial, industrial, and other intangible assets rental and leasing \tab miscellaneous petroleum and coal products \tab other administrative and other support services \tab other motor vehicle dealers \tab 149 \tab 1720 \tab music stores \tab data processing, hosting, and related services \tab offices of other health practitioners \tab other health care services \tab sound recording industries \tab travel arrangements and reservation services \tab child day care services \tab libraries and archives \tab recreational vehicle parks and camps, and rooming and boarding houses \tab outpatient care centers \tab vocational rehabilitation services \tab gift, novelty, and souvenir shops \tab nail salons and other personal care services \tab car washes \tab home health care services \tab community food and housing, and emergency services \tab 2320 \tab 9720 \tab software publishers \tab Total \cr
   1972 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 2 \tab 5 \tab 4 \tab 6 \tab 56 \tab 5 \tab 3 \tab 2 \tab 1 \tab 8 \tab 12 \tab 3 \tab 8 \tab 12 \tab 7 \tab 10 \tab 7 \tab 1 \tab 6 \tab 32 \tab 1 \tab 2 \tab 3 \tab 2 \tab 3 \tab 133 \tab 1 \tab 89 \tab 6 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 8 \tab 8 \tab 1 \tab 19 \tab 1 \tab 2 \tab 6 \tab 6 \tab 4 \tab 2 \tab 1 \tab 3 \tab 2 \tab 2 \tab 3 \tab 2 \tab 1 \tab 4 \tab 8 \tab 1 \tab 11 \tab 12 \tab 1 \tab 1 \tab 1 \tab 7 \tab 3 \tab 1 \tab 198 \tab 1 \tab 1 \tab 12 \tab 14 \tab 1 \tab 19 \tab 4 \tab 27 \tab 2 \tab 1 \tab 1 \tab 5 \tab 2 \tab 7 \tab 2 \tab 6 \tab 2 \tab 2 \tab 8 \tab 1 \tab 3 \tab 1 \tab 7 \tab 8 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 14 \tab 1 \tab 20 \tab 1 \tab 1 \tab 2 \tab 75 \tab 1 \tab 6 \tab 2 \tab 1 \tab 4 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 6 \tab 4 \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 6 \tab 2 \tab 5 \tab 2 \tab 4 \tab 1 \tab 15 \tab 1 \tab 4 \tab 2 \tab 3 \tab 2 \tab 43 \tab 2 \tab 11 \tab 3 \tab 12 \tab 3 \tab 3 \tab 2 \tab 17 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 8 \tab 2 \tab 2 \tab 4 \tab 1 \tab 2 \tab 2 \tab 1 \tab 6 \tab 22 \tab 1 \tab 1 \tab 1 \tab 288 \tab 4 \tab 1 \tab 6 \tab 4 \tab 1 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 4 \tab - \tab - \tab - \tab - \tab 2 \tab 6 \tab 1 \tab 3 \tab 98 \tab 3 \tab 1 \tab 5 \tab - \tab 9 \tab 6 \tab 3 \tab 8 \tab 6 \tab 1 \tab 8 \tab 4 \tab - \tab 4 \tab 42 \tab 5 \tab 6 \tab 2 \tab 2 \tab - \tab 136 \tab 3 \tab 163 \tab 4 \tab 1 \tab 5 \tab 1 \tab - \tab 4 \tab 2 \tab 3 \tab 9 \tab 3 \tab - \tab 8 \tab 2 \tab 3 \tab 4 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 5 \tab 3 \tab 2 \tab 4 \tab 7 \tab 5 \tab 5 \tab 15 \tab 21 \tab 5 \tab 1 \tab 1 \tab 5 \tab 5 \tab 6 \tab 151 \tab 3 \tab - \tab 7 \tab 12 \tab 1 \tab 27 \tab 3 \tab 19 \tab 2 \tab 1 \tab 3 \tab - \tab 2 \tab 3 \tab 4 \tab 10 \tab 2 \tab - \tab 8 \tab 1 \tab 3 \tab - \tab 8 \tab 9 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 20 \tab 1 \tab 12 \tab - \tab - \tab 6 \tab 28 \tab 2 \tab 3 \tab 2 \tab - \tab 2 \tab 6 \tab 3 \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 2 \tab 7 \tab 1 \tab 9 \tab - \tab 11 \tab - \tab 13 \tab - \tab 64 \tab 2 \tab 7 \tab 4 \tab 13 \tab 6 \tab 4 \tab 3 \tab 21 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 8 \tab 1 \tab - \tab 2 \tab - \tab 4 \tab 8 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 21 \tab 2 \tab 1 \tab 2 \tab 111 \tab 2 \tab 1 \tab 2 \tab 5 \tab 2 \tab 4 \tab 1 \tab 4 \tab 1 \tab 4 \tab 4 \tab 1 \tab 1 \tab 1 \tab 6 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab - \tab - \tab 90 \tab 3 \tab 1 \tab 3 \tab 3 \tab 11 \tab 10 \tab 1 \tab 2 \tab 10 \tab 7 \tab 5 \tab 10 \tab 1 \tab 4 \tab 36 \tab 5 \tab 2 \tab 3 \tab - \tab - \tab 130 \tab 1 \tab 114 \tab 3 \tab - \tab 6 \tab 2 \tab - \tab 2 \tab 1 \tab 5 \tab 4 \tab 8 \tab 1 \tab 12 \tab 3 \tab 4 \tab 7 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 5 \tab 2 \tab 1 \tab 3 \tab 8 \tab 5 \tab 4 \tab 9 \tab 17 \tab 5 \tab 3 \tab - \tab 8 \tab 3 \tab 1 \tab 182 \tab - \tab 1 \tab 10 \tab 9 \tab - \tab 22 \tab 2 \tab 11 \tab - \tab - \tab - \tab 1 \tab 3 \tab 9 \tab - \tab 10 \tab 3 \tab - \tab 11 \tab 1 \tab 3 \tab - \tab 6 \tab 7 \tab 8 \tab 2 \tab 3 \tab 3 \tab 6 \tab 22 \tab 1 \tab 11 \tab 1 \tab 1 \tab 8 \tab 92 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab 3 \tab - \tab - \tab 5 \tab 2 \tab 2 \tab 7 \tab - \tab 1 \tab 5 \tab 6 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 4 \tab 4 \tab 2 \tab 3 \tab 3 \tab 19 \tab 2 \tab 2 \tab 5 \tab 2 \tab - \tab 52 \tab 2 \tab 8 \tab - \tab 8 \tab 9 \tab 4 \tab - \tab 15 \tab - \tab 5 \tab 6 \tab - \tab 1 \tab 6 \tab 3 \tab 1 \tab 5 \tab - \tab 3 \tab 1 \tab - \tab 3 \tab 3 \tab 4 \tab 4 \tab 12 \tab 2 \tab 1 \tab 3 \tab 87 \tab 2 \tab - \tab 2 \tab 7 \tab - \tab 7 \tab 2 \tab 5 \tab 2 \tab 6 \tab 2 \tab - \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 6 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 2 \tab 4 \tab 1 \tab 2 \tab - \tab 2 \tab 15 \tab - \tab 4 \tab 99 \tab 6 \tab 2 \tab 3 \tab 1 \tab 14 \tab 13 \tab 1 \tab 5 \tab 10 \tab 7 \tab 5 \tab 6 \tab - \tab 4 \tab 34 \tab 8 \tab 2 \tab 2 \tab 2 \tab 2 \tab 131 \tab 4 \tab 121 \tab 3 \tab 2 \tab 11 \tab 6 \tab - \tab 1 \tab 2 \tab 5 \tab 10 \tab 4 \tab - \tab 10 \tab 1 \tab 6 \tab 6 \tab 2 \tab 1 \tab 5 \tab 2 \tab 2 \tab 3 \tab 2 \tab 4 \tab - \tab 4 \tab 10 \tab 10 \tab 6 \tab 4 \tab 23 \tab 2 \tab 3 \tab - \tab 4 \tab 5 \tab 1 \tab 153 \tab 1 \tab - \tab 7 \tab 13 \tab 1 \tab 25 \tab 2 \tab 19 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 11 \tab 2 \tab 1 \tab 15 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab 5 \tab 1 \tab 2 \tab - \tab 3 \tab 25 \tab 1 \tab 23 \tab 3 \tab - \tab 3 \tab 18 \tab 2 \tab 6 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 7 \tab - \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 5 \tab 3 \tab 3 \tab 2 \tab 6 \tab - \tab 16 \tab 3 \tab 5 \tab 3 \tab 3 \tab 2 \tab 43 \tab 3 \tab 12 \tab - \tab 8 \tab 6 \tab - \tab 1 \tab 20 \tab 1 \tab - \tab 10 \tab 1 \tab - \tab 10 \tab 1 \tab - \tab 5 \tab 3 \tab 4 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 6 \tab 26 \tab 3 \tab 1 \tab 3 \tab 119 \tab 2 \tab 4 \tab 3 \tab 4 \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 2 \tab 4 \tab - \tab - \tab - \tab 5 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 6 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 2 \tab - \tab 3 \tab 2 \tab 5 \tab - \tab 17 \tab 1 \tab 2 \tab 114 \tab 2 \tab - \tab 2 \tab 1 \tab 11 \tab 13 \tab 3 \tab 7 \tab 2 \tab 4 \tab 11 \tab 4 \tab - \tab 4 \tab 48 \tab 4 \tab 2 \tab 2 \tab 4 \tab 1 \tab 111 \tab 4 \tab 127 \tab 7 \tab 3 \tab 9 \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab 7 \tab 4 \tab 2 \tab 15 \tab 1 \tab 5 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab - \tab 10 \tab 4 \tab 2 \tab 3 \tab 4 \tab 3 \tab 7 \tab 12 \tab 13 \tab 3 \tab 2 \tab - \tab 7 \tab 3 \tab 1 \tab 187 \tab 5 \tab - \tab 6 \tab 16 \tab - \tab 39 \tab 2 \tab 16 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 3 \tab 1 \tab 10 \tab 1 \tab 1 \tab 15 \tab 2 \tab 8 \tab 2 \tab 6 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 6 \tab 35 \tab 1 \tab 15 \tab 4 \tab 1 \tab 4 \tab 24 \tab 2 \tab 2 \tab 4 \tab - \tab 2 \tab 6 \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 2 \tab 10 \tab 2 \tab 2 \tab - \tab 5 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 3 \tab 4 \tab 2 \tab 2 \tab - \tab 3 \tab 7 \tab 1 \tab 6 \tab 4 \tab 4 \tab 2 \tab 51 \tab 3 \tab 10 \tab 3 \tab 9 \tab 10 \tab 3 \tab - \tab 12 \tab 2 \tab 1 \tab 5 \tab 1 \tab 2 \tab 4 \tab 1 \tab - \tab 4 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 13 \tab 3 \tab - \tab 2 \tab 76 \tab 4 \tab 3 \tab 3 \tab 2 \tab 1 \tab 9 \tab - \tab 2 \tab 2 \tab 6 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 3 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 4 \tab 10 \tab 2 \tab 3 \tab 58 \tab - \tab - \tab 4 \tab 3 \tab 4 \tab 13 \tab 2 \tab 7 \tab 3 \tab 8 \tab 4 \tab 4 \tab 2 \tab 5 \tab 56 \tab 7 \tab 5 \tab 3 \tab 4 \tab 2 \tab 123 \tab 5 \tab 108 \tab - \tab 1 \tab 7 \tab 3 \tab - \tab 3 \tab 1 \tab 9 \tab 12 \tab 5 \tab - \tab 12 \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab - \tab 10 \tab 2 \tab - \tab 3 \tab 7 \tab 5 \tab 2 \tab 10 \tab 16 \tab 3 \tab 3 \tab - \tab 2 \tab 3 \tab 1 \tab 193 \tab 3 \tab - \tab 3 \tab 9 \tab - \tab 28 \tab 2 \tab 20 \tab - \tab - \tab 2 \tab 5 \tab 8 \tab 6 \tab 2 \tab 13 \tab 1 \tab 1 \tab 17 \tab - \tab 7 \tab - \tab 4 \tab 2 \tab 5 \tab 3 \tab 1 \tab 3 \tab 3 \tab 29 \tab 1 \tab 26 \tab - \tab 2 \tab 6 \tab 25 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 5 \tab 2 \tab - \tab - \tab 3 \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 5 \tab 1 \tab 6 \tab - \tab 2 \tab 2 \tab 18 \tab 3 \tab 1 \tab 5 \tab 7 \tab 1 \tab 45 \tab 2 \tab 5 \tab 3 \tab 8 \tab 4 \tab 2 \tab 2 \tab 18 \tab 2 \tab 4 \tab 9 \tab 2 \tab 1 \tab 10 \tab 1 \tab - \tab 3 \tab 4 \tab 8 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 8 \tab 22 \tab - \tab 4 \tab 2 \tab 186 \tab 1 \tab - \tab 5 \tab 6 \tab 1 \tab 5 \tab - \tab 1 \tab 2 \tab 6 \tab 3 \tab - \tab - \tab - \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 3 \tab 15 \tab 2 \tab 5 \tab 126 \tab 5 \tab - \tab 5 \tab 1 \tab 9 \tab 7 \tab 3 \tab 5 \tab 3 \tab 3 \tab 5 \tab 6 \tab - \tab 3 \tab 56 \tab 3 \tab 4 \tab 5 \tab 1 \tab 2 \tab 141 \tab 3 \tab 139 \tab 4 \tab 4 \tab 5 \tab 7 \tab 2 \tab 4 \tab - \tab 6 \tab 12 \tab 3 \tab 2 \tab 14 \tab 1 \tab 8 \tab 11 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 4 \tab 9 \tab 2 \tab 1 \tab 1 \tab 12 \tab 8 \tab 8 \tab 16 \tab 16 \tab 1 \tab 2 \tab - \tab 6 \tab - \tab 2 \tab 205 \tab 3 \tab 1 \tab 9 \tab 13 \tab - \tab 26 \tab 3 \tab 29 \tab 1 \tab - \tab 2 \tab 2 \tab 9 \tab 9 \tab 2 \tab 3 \tab 3 \tab 1 \tab 17 \tab - \tab 3 \tab - \tab 8 \tab 3 \tab 5 \tab 1 \tab 1 \tab - \tab 2 \tab 31 \tab - \tab 15 \tab 3 \tab 2 \tab 5 \tab 17 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 5 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 6 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 4 \tab 3 \tab - \tab 3 \tab 1 \tab 15 \tab 4 \tab 8 \tab 3 \tab 5 \tab 3 \tab 44 \tab - \tab 4 \tab 5 \tab 7 \tab 11 \tab 8 \tab 2 \tab 11 \tab 2 \tab 2 \tab 5 \tab 1 \tab 1 \tab 9 \tab 3 \tab 2 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 5 \tab 3 \tab 2 \tab 16 \tab 8 \tab 7 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 2 \tab 5 \tab 2 \tab 4 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab 4 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab 3 \tab 2 \tab 1 \tab 2 \tab 8 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 2 \tab 4 \tab - \tab 4 \tab 2 \tab 2 \tab 6 \tab 2 \tab 3 \tab 96 \tab 4 \tab 3 \tab 4 \tab 3 \tab 11 \tab 12 \tab 3 \tab 5 \tab 9 \tab 4 \tab 6 \tab 3 \tab 1 \tab 4 \tab 55 \tab 10 \tab 3 \tab 4 \tab 2 \tab - \tab 111 \tab 3 \tab 120 \tab 3 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab 4 \tab 10 \tab 1 \tab - \tab 19 \tab 3 \tab 5 \tab 7 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 4 \tab 2 \tab 2 \tab 5 \tab 8 \tab 8 \tab 7 \tab 10 \tab 23 \tab 2 \tab 1 \tab - \tab 7 \tab 1 \tab 2 \tab 230 \tab - \tab - \tab 9 \tab 10 \tab - \tab 28 \tab 3 \tab 21 \tab 2 \tab 2 \tab 3 \tab 3 \tab 7 \tab 7 \tab 2 \tab 8 \tab - \tab - \tab 16 \tab 1 \tab 2 \tab 1 \tab 4 \tab 5 \tab 6 \tab 2 \tab 1 \tab 1 \tab 3 \tab 16 \tab 2 \tab 15 \tab 4 \tab 1 \tab 9 \tab 9 \tab - \tab 2 \tab 4 \tab 3 \tab - \tab 8 \tab - \tab 5 \tab 1 \tab - \tab 2 \tab - \tab 8 \tab - \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 6 \tab 3 \tab 2 \tab 2 \tab 2 \tab 2 \tab 9 \tab 1 \tab 9 \tab 4 \tab 7 \tab - \tab 44 \tab 4 \tab 9 \tab 1 \tab 6 \tab 8 \tab 2 \tab 2 \tab 15 \tab 3 \tab 3 \tab 3 \tab 2 \tab 1 \tab 4 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab 3 \tab - \tab 2 \tab 3 \tab - \tab 7 \tab 18 \tab 5 \tab 4 \tab 3 \tab 46 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 10 \tab 1 \tab 3 \tab 1 \tab 4 \tab 3 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab 3 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab 8 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 3 \tab 1 \tab 4 \tab - \tab 2 \tab 5 \tab 4 \tab 3 \tab 82 \tab 4 \tab 2 \tab 4 \tab 4 \tab 11 \tab 19 \tab 2 \tab 7 \tab 8 \tab 10 \tab 3 \tab 11 \tab - \tab 3 \tab 44 \tab 5 \tab 1 \tab 7 \tab 2 \tab 1 \tab 167 \tab 3 \tab 188 \tab 9 \tab - \tab 5 \tab 6 \tab 2 \tab 5 \tab 2 \tab 6 \tab 9 \tab 5 \tab 1 \tab 25 \tab 2 \tab 9 \tab 3 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 3 \tab 10 \tab - \tab 1 \tab 4 \tab 6 \tab 9 \tab 4 \tab 11 \tab 26 \tab 5 \tab 1 \tab - \tab 11 \tab 2 \tab 1 \tab 296 \tab 2 \tab 1 \tab 12 \tab 10 \tab 1 \tab 35 \tab 2 \tab 19 \tab 2 \tab 3 \tab 4 \tab 3 \tab 3 \tab 11 \tab 1 \tab 9 \tab 5 \tab - \tab 12 \tab - \tab 5 \tab 2 \tab 5 \tab 5 \tab 5 \tab - \tab - \tab - \tab 1 \tab 23 \tab 4 \tab 22 \tab 4 \tab 1 \tab 8 \tab 85 \tab 2 \tab 5 \tab 1 \tab 3 \tab - \tab 9 \tab 3 \tab 7 \tab 6 \tab - \tab - \tab 3 \tab 3 \tab 3 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab 4 \tab - \tab 2 \tab 4 \tab 1 \tab 6 \tab 6 \tab - \tab 1 \tab - \tab 19 \tab 4 \tab 4 \tab 3 \tab 4 \tab - \tab 46 \tab 1 \tab 12 \tab 1 \tab 13 \tab 13 \tab 5 \tab - \tab 21 \tab 1 \tab 1 \tab 9 \tab 2 \tab - \tab 10 \tab 1 \tab 1 \tab 4 \tab 4 \tab 8 \tab 2 \tab 1 \tab 1 \tab 6 \tab 1 \tab 4 \tab 33 \tab 4 \tab 3 \tab 3 \tab 20 \tab 4 \tab 4 \tab 4 \tab 3 \tab 2 \tab 4 \tab - \tab 4 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 4 \tab 4 \tab 5 \tab 1 \tab - \tab 5 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 2 \tab 1 \tab 2 \tab 8 \tab - \tab 2 \tab 8 \tab 1 \tab 3 \tab 71 \tab 11 \tab 1 \tab 6 \tab - \tab 16 \tab 16 \tab 4 \tab 10 \tab 10 \tab 1 \tab 7 \tab 6 \tab - \tab 7 \tab 18 \tab 11 \tab 4 \tab 2 \tab 6 \tab 2 \tab 170 \tab 3 \tab 106 \tab 7 \tab 1 \tab 10 \tab 10 \tab - \tab 1 \tab 1 \tab 6 \tab 15 \tab 4 \tab 3 \tab 23 \tab 2 \tab 6 \tab 10 \tab 2 \tab 1 \tab 3 \tab 1 \tab 5 \tab - \tab 8 \tab 4 \tab 1 \tab 7 \tab 15 \tab 7 \tab 2 \tab 6 \tab 20 \tab 2 \tab 5 \tab - \tab 6 \tab 2 \tab 3 \tab 207 \tab - \tab 1 \tab 12 \tab 12 \tab - \tab 56 \tab 1 \tab 21 \tab 6 \tab 4 \tab 5 \tab 2 \tab 9 \tab 6 \tab 1 \tab 3 \tab 3 \tab 1 \tab 18 \tab 3 \tab 7 \tab - \tab 5 \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab 27 \tab - \tab 21 \tab 1 \tab 6 \tab 4 \tab 18 \tab 1 \tab - \tab 6 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 6 \tab 8 \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 5 \tab - \tab 4 \tab 2 \tab 1 \tab 3 \tab 19 \tab - \tab 4 \tab 3 \tab 6 \tab 1 \tab 35 \tab 1 \tab 12 \tab 2 \tab 15 \tab 8 \tab 3 \tab 2 \tab 13 \tab 4 \tab 7 \tab 7 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 5 \tab 2 \tab 11 \tab 2 \tab - \tab 1 \tab 4 \tab 3 \tab 8 \tab 26 \tab 3 \tab 4 \tab - \tab 23 \tab 5 \tab 3 \tab 5 \tab 5 \tab 2 \tab 8 \tab - \tab 3 \tab 1 \tab 8 \tab 5 \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 5 \tab - \tab - \tab 3 \tab 3 \tab 4 \tab 1 \tab 3 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 5 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 2 \tab 1 \tab - \tab 4 \tab - \tab 4 \tab 10 \tab - \tab 3 \tab 126 \tab 8 \tab - \tab 8 \tab 4 \tab 11 \tab 11 \tab 2 \tab 2 \tab 8 \tab 5 \tab 8 \tab 5 \tab - \tab 3 \tab 28 \tab 8 \tab 6 \tab - \tab 2 \tab 3 \tab 140 \tab 1 \tab 127 \tab 9 \tab 1 \tab 8 \tab 4 \tab - \tab 5 \tab 1 \tab 4 \tab 11 \tab 8 \tab 1 \tab 19 \tab 2 \tab 4 \tab 9 \tab 1 \tab - \tab 5 \tab - \tab 3 \tab 2 \tab 7 \tab 1 \tab 2 \tab 5 \tab 6 \tab 8 \tab 8 \tab 9 \tab 21 \tab 2 \tab - \tab 2 \tab 7 \tab 1 \tab - \tab 202 \tab - \tab 1 \tab 7 \tab 19 \tab 2 \tab 29 \tab 3 \tab 15 \tab 2 \tab 3 \tab 4 \tab 1 \tab 7 \tab 3 \tab - \tab 8 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 3 \tab 3 \tab 5 \tab 4 \tab 4 \tab - \tab 3 \tab 29 \tab 1 \tab 15 \tab 3 \tab 2 \tab 7 \tab 18 \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab 5 \tab 4 \tab 3 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 12 \tab - \tab 7 \tab 1 \tab 7 \tab 1 \tab 31 \tab 4 \tab 9 \tab 2 \tab 8 \tab 13 \tab - \tab 1 \tab 8 \tab 2 \tab 3 \tab 12 \tab 2 \tab 1 \tab 9 \tab - \tab - \tab 6 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 22 \tab 3 \tab 8 \tab - \tab 2 \tab 3 \tab 1 \tab 4 \tab 3 \tab - \tab 12 \tab - \tab 3 \tab 1 \tab 7 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab 6 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 8 \tab 1 \tab 4 \tab 110 \tab 13 \tab 3 \tab 5 \tab 1 \tab 15 \tab 14 \tab 4 \tab 5 \tab 4 \tab 5 \tab 4 \tab 3 \tab 1 \tab 3 \tab 19 \tab 8 \tab 1 \tab 2 \tab 2 \tab 3 \tab 148 \tab 3 \tab 110 \tab 3 \tab - \tab 1 \tab 4 \tab - \tab 4 \tab 1 \tab 3 \tab 9 \tab 5 \tab 2 \tab 18 \tab 4 \tab 7 \tab 12 \tab 1 \tab - \tab 2 \tab - \tab 4 \tab - \tab 6 \tab 5 \tab 1 \tab 4 \tab 6 \tab 8 \tab 6 \tab 5 \tab 26 \tab 1 \tab 2 \tab - \tab 8 \tab 1 \tab 2 \tab 194 \tab 1 \tab 3 \tab 7 \tab 19 \tab - \tab 41 \tab - \tab 19 \tab 2 \tab 2 \tab 2 \tab 1 \tab 8 \tab 2 \tab 2 \tab 9 \tab 2 \tab 1 \tab 5 \tab - \tab 3 \tab 1 \tab 4 \tab 4 \tab 5 \tab 3 \tab 1 \tab 1 \tab 5 \tab 26 \tab 2 \tab 26 \tab 5 \tab 6 \tab 2 \tab 8 \tab 2 \tab 2 \tab 4 \tab 1 \tab 3 \tab 10 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 3 \tab 11 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab 4 \tab 2 \tab 4 \tab 2 \tab 5 \tab - \tab 19 \tab 4 \tab 6 \tab 3 \tab 10 \tab 1 \tab 38 \tab - \tab 10 \tab 3 \tab 13 \tab 11 \tab 5 \tab - \tab 10 \tab 4 \tab 3 \tab 5 \tab 4 \tab 2 \tab 7 \tab 1 \tab 1 \tab 8 \tab 5 \tab 7 \tab 3 \tab - \tab 1 \tab 6 \tab 1 \tab 2 \tab 23 \tab 5 \tab 5 \tab 4 \tab 55 \tab 2 \tab 3 \tab 2 \tab 6 \tab 4 \tab 8 \tab - \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 6 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 8 \tab 1 \tab 5 \tab 124 \tab 10 \tab 3 \tab 2 \tab 3 \tab 15 \tab 10 \tab 1 \tab 4 \tab 4 \tab 9 \tab 9 \tab 8 \tab 1 \tab - \tab 10 \tab 10 \tab 6 \tab 2 \tab 2 \tab 1 \tab 124 \tab 4 \tab 105 \tab 11 \tab - \tab 2 \tab 9 \tab - \tab 2 \tab 3 \tab 7 \tab 10 \tab 7 \tab 2 \tab 11 \tab - \tab 6 \tab 8 \tab 2 \tab 2 \tab - \tab 5 \tab 1 \tab 1 \tab 9 \tab 6 \tab 3 \tab 5 \tab 12 \tab 5 \tab 7 \tab 10 \tab 15 \tab 3 \tab 1 \tab - \tab 10 \tab 2 \tab 1 \tab 222 \tab 6 \tab - \tab 10 \tab 12 \tab - \tab 39 \tab 1 \tab 28 \tab 3 \tab 6 \tab 2 \tab 1 \tab 5 \tab 3 \tab 4 \tab 11 \tab 4 \tab 1 \tab 6 \tab - \tab 2 \tab - \tab 3 \tab 3 \tab 7 \tab 5 \tab 4 \tab 2 \tab 2 \tab 28 \tab 1 \tab 13 \tab 5 \tab 1 \tab 5 \tab 21 \tab 1 \tab 3 \tab 5 \tab 1 \tab 1 \tab 10 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab 3 \tab 3 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 5 \tab 2 \tab - \tab 3 \tab 3 \tab 1 \tab 3 \tab - \tab 12 \tab - \tab 6 \tab 5 \tab 5 \tab 3 \tab 32 \tab - \tab 13 \tab 1 \tab 10 \tab 9 \tab 2 \tab - \tab 7 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 5 \tab 1 \tab 3 \tab - \tab 2 \tab 5 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 4 \tab 23 \tab 2 \tab 2 \tab 3 \tab 10 \tab 4 \tab 3 \tab 4 \tab 2 \tab - \tab 6 \tab - \tab 4 \tab 4 \tab 3 \tab 4 \tab - \tab 2 \tab - \tab 5 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 4 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 3 \tab 2 \tab 5 \tab 4 \tab 2 \tab 6 \tab 9 \tab 4 \tab 5 \tab 117 \tab 8 \tab 1 \tab 7 \tab 4 \tab 13 \tab 11 \tab 4 \tab 6 \tab 2 \tab 1 \tab 3 \tab 12 \tab 1 \tab 1 \tab 28 \tab 13 \tab 5 \tab 1 \tab - \tab 1 \tab 168 \tab 3 \tab 139 \tab 6 \tab 2 \tab 5 \tab 7 \tab - \tab 2 \tab - \tab 6 \tab 10 \tab 5 \tab 2 \tab 24 \tab 2 \tab 6 \tab 11 \tab 2 \tab 1 \tab 2 \tab 2 \tab 11 \tab 1 \tab 6 \tab 3 \tab - \tab 6 \tab 6 \tab 4 \tab 7 \tab 9 \tab 17 \tab 2 \tab 1 \tab - \tab 13 \tab 6 \tab 2 \tab 291 \tab 3 \tab - \tab 13 \tab 17 \tab 1 \tab 39 \tab - \tab 22 \tab - \tab 2 \tab 1 \tab 4 \tab 8 \tab 7 \tab 4 \tab 20 \tab 3 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab 4 \tab 5 \tab 7 \tab - \tab 4 \tab 3 \tab 2 \tab 38 \tab 2 \tab 20 \tab 6 \tab 1 \tab 3 \tab 29 \tab 1 \tab 4 \tab 3 \tab 1 \tab 2 \tab 10 \tab 4 \tab 2 \tab 5 \tab - \tab - \tab 2 \tab 9 \tab 3 \tab 3 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 5 \tab 3 \tab 7 \tab 3 \tab 3 \tab 2 \tab 2 \tab 2 \tab 11 \tab 1 \tab 8 \tab 3 \tab 10 \tab 2 \tab 47 \tab 1 \tab 9 \tab 2 \tab 13 \tab 8 \tab 3 \tab 1 \tab 7 \tab - \tab 3 \tab - \tab 1 \tab - \tab 8 \tab - \tab - \tab 5 \tab 5 \tab 4 \tab 6 \tab - \tab - \tab 8 \tab 2 \tab 4 \tab 27 \tab 3 \tab 8 \tab - \tab 82 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 4 \tab 1 \tab 3 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab 5 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab 5 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 4 \tab 2 \tab - \tab 6 \tab 1 \tab 2 \tab 9 \tab - \tab 2 \tab 102 \tab 7 \tab 1 \tab 8 \tab 2 \tab 13 \tab 15 \tab 3 \tab 1 \tab 4 \tab 3 \tab 10 \tab 11 \tab - \tab 3 \tab 12 \tab 14 \tab 1 \tab 1 \tab 3 \tab 1 \tab 130 \tab 2 \tab 119 \tab 3 \tab 2 \tab 5 \tab 3 \tab 2 \tab 3 \tab - \tab 2 \tab 8 \tab 4 \tab 3 \tab 28 \tab 1 \tab 16 \tab 14 \tab 5 \tab 1 \tab 1 \tab 2 \tab 5 \tab 2 \tab 8 \tab 2 \tab 1 \tab 5 \tab 9 \tab 9 \tab 1 \tab 7 \tab 21 \tab 7 \tab 1 \tab - \tab 10 \tab 1 \tab - \tab 207 \tab 4 \tab - \tab 8 \tab 10 \tab 1 \tab 19 \tab 4 \tab 23 \tab - \tab 3 \tab 3 \tab 1 \tab 5 \tab 8 \tab 2 \tab 12 \tab 2 \tab 2 \tab 4 \tab 1 \tab 2 \tab - \tab 5 \tab 5 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 32 \tab 3 \tab 19 \tab 5 \tab 3 \tab 6 \tab 14 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 15 \tab 2 \tab 7 \tab 2 \tab 1 \tab - \tab 2 \tab 6 \tab 4 \tab 3 \tab - \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab 2 \tab 9 \tab 2 \tab 3 \tab 3 \tab 2 \tab 8 \tab 4 \tab 4 \tab 1 \tab 6 \tab 5 \tab 32 \tab 1 \tab 9 \tab 1 \tab 11 \tab 9 \tab 4 \tab 3 \tab 11 \tab 4 \tab 1 \tab 10 \tab 3 \tab 3 \tab 6 \tab 1 \tab - \tab 2 \tab 1 \tab 5 \tab 7 \tab 2 \tab - \tab 2 \tab - \tab 5 \tab 19 \tab 6 \tab 5 \tab 4 \tab - \tab 3 \tab 3 \tab 1 \tab 1 \tab 3 \tab 10 \tab - \tab 7 \tab 4 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 4 \tab 1 \tab 5 \tab 1 \tab 4 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 4 \tab 4 \tab - \tab 1 \tab 3 \tab 4 \tab 12 \tab - \tab 5 \tab 125 \tab 14 \tab - \tab 8 \tab 4 \tab 13 \tab 20 \tab 3 \tab 4 \tab 6 \tab 4 \tab 7 \tab 9 \tab - \tab 1 \tab 21 \tab 14 \tab 4 \tab - \tab 2 \tab 3 \tab 124 \tab 5 \tab 113 \tab 5 \tab 1 \tab 2 \tab 6 \tab 1 \tab 4 \tab 1 \tab 7 \tab 12 \tab 8 \tab 2 \tab 25 \tab 2 \tab 4 \tab 9 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab 4 \tab 6 \tab 2 \tab 5 \tab 17 \tab 13 \tab 4 \tab - \tab - \tab 5 \tab 2 \tab 1 \tab 225 \tab 2 \tab 1 \tab 5 \tab 13 \tab - \tab 36 \tab 1 \tab 23 \tab - \tab 6 \tab 4 \tab 3 \tab 6 \tab 6 \tab 2 \tab 4 \tab 1 \tab - \tab 3 \tab - \tab 6 \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab 33 \tab - \tab 23 \tab 2 \tab 3 \tab 6 \tab 16 \tab 2 \tab 2 \tab 8 \tab 1 \tab 2 \tab 4 \tab 1 \tab 5 \tab 4 \tab 1 \tab 1 \tab - \tab 5 \tab 4 \tab 6 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 5 \tab 3 \tab 2 \tab 5 \tab 2 \tab 1 \tab 4 \tab 9 \tab 22 \tab 1 \tab 4 \tab 4 \tab 7 \tab 1 \tab 32 \tab 3 \tab 16 \tab 3 \tab 12 \tab 12 \tab 3 \tab - \tab 8 \tab 3 \tab 3 \tab 6 \tab 2 \tab - \tab 5 \tab 2 \tab 2 \tab 3 \tab - \tab 5 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 4 \tab 39 \tab 7 \tab 2 \tab 5 \tab 6 \tab - \tab 4 \tab 5 \tab 4 \tab 1 \tab 9 \tab - \tab 5 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab 4 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 2 \tab 1 \tab 4 \tab 3 \tab 1 \tab 4 \tab 15 \tab - \tab 4 \tab 88 \tab 9 \tab 3 \tab 3 \tab 3 \tab 10 \tab 15 \tab 2 \tab 6 \tab 3 \tab 5 \tab 10 \tab 7 \tab - \tab 4 \tab 9 \tab 8 \tab 3 \tab 1 \tab 2 \tab 1 \tab 122 \tab 4 \tab 89 \tab 5 \tab 2 \tab 2 \tab 6 \tab 1 \tab 1 \tab 2 \tab 11 \tab 10 \tab 7 \tab 1 \tab 18 \tab 2 \tab 10 \tab 4 \tab 4 \tab - \tab - \tab 2 \tab 4 \tab - \tab 4 \tab 5 \tab 2 \tab 4 \tab 11 \tab 6 \tab 6 \tab 15 \tab 15 \tab 2 \tab 5 \tab 1 \tab 6 \tab 3 \tab - \tab 217 \tab 4 \tab 1 \tab 9 \tab 14 \tab - \tab 37 \tab 1 \tab 11 \tab 2 \tab 2 \tab 5 \tab - \tab - \tab 3 \tab 1 \tab 5 \tab 3 \tab 1 \tab 5 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 39 \tab - \tab 13 \tab 6 \tab 3 \tab 7 \tab 16 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 6 \tab 2 \tab 6 \tab 4 \tab - \tab 1 \tab 3 \tab 4 \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 3 \tab 7 \tab 1 \tab 1 \tab 3 \tab 15 \tab 1 \tab 4 \tab - \tab 7 \tab 2 \tab 32 \tab 1 \tab 4 \tab 4 \tab 8 \tab 9 \tab 2 \tab - \tab 6 \tab 1 \tab 3 \tab 6 \tab 3 \tab - \tab 5 \tab - \tab - \tab 6 \tab - \tab 4 \tab 1 \tab - \tab - \tab 3 \tab - \tab 2 \tab 16 \tab 1 \tab 6 \tab 3 \tab - \tab 3 \tab 2 \tab 5 \tab 4 \tab 1 \tab 15 \tab 1 \tab 5 \tab 5 \tab 5 \tab 5 \tab 2 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab 4 \tab 5 \tab 5 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 2 \tab 2 \tab 5 \tab 6 \tab - \tab 2 \tab 10 \tab - \tab 4 \tab 110 \tab 9 \tab 2 \tab 8 \tab 4 \tab 12 \tab 16 \tab 3 \tab 3 \tab 6 \tab 4 \tab 5 \tab 9 \tab 2 \tab 2 \tab 23 \tab 7 \tab 4 \tab 2 \tab - \tab 2 \tab 132 \tab 2 \tab 105 \tab 7 \tab - \tab 4 \tab 7 \tab 3 \tab 2 \tab - \tab 5 \tab 10 \tab 2 \tab - \tab 24 \tab 3 \tab 6 \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 7 \tab 2 \tab 3 \tab 7 \tab 6 \tab 6 \tab 4 \tab 8 \tab 18 \tab 3 \tab 1 \tab - \tab 12 \tab 1 \tab 1 \tab 220 \tab 3 \tab 1 \tab 7 \tab 16 \tab - \tab 29 \tab - \tab 7 \tab 2 \tab 4 \tab 4 \tab 1 \tab 2 \tab 4 \tab 2 \tab 10 \tab - \tab - \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab 5 \tab 5 \tab 1 \tab - \tab 1 \tab 3 \tab 33 \tab 1 \tab 26 \tab 3 \tab 4 \tab 6 \tab 44 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 5 \tab - \tab 6 \tab 7 \tab 1 \tab 1 \tab 4 \tab 8 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 5 \tab 2 \tab 3 \tab 4 \tab 3 \tab - \tab 21 \tab 3 \tab 7 \tab 2 \tab 9 \tab 1 \tab 30 \tab 2 \tab 6 \tab - \tab 12 \tab 10 \tab 4 \tab 2 \tab 8 \tab 3 \tab 4 \tab 8 \tab 3 \tab 1 \tab 6 \tab 1 \tab 3 \tab 3 \tab - \tab 8 \tab 2 \tab - \tab - \tab 3 \tab 3 \tab 4 \tab 27 \tab 7 \tab 9 \tab 1 \tab 1 \tab 4 \tab 4 \tab 1 \tab 5 \tab 3 \tab 8 \tab - \tab 7 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 3 \tab 4 \tab 2 \tab 4 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1 \tab 4 \tab 4 \tab 11 \tab 1 \tab 6 \tab 6 \tab 4 \tab 3 \tab 106 \tab 4 \tab - \tab 7 \tab 4 \tab 19 \tab 20 \tab 2 \tab 9 \tab 3 \tab 1 \tab 6 \tab 7 \tab - \tab 4 \tab 10 \tab 10 \tab 7 \tab 5 \tab 3 \tab 5 \tab 130 \tab 4 \tab 126 \tab 7 \tab 1 \tab 4 \tab 5 \tab - \tab 2 \tab 3 \tab 4 \tab 20 \tab 2 \tab 3 \tab 23 \tab - \tab 12 \tab 9 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab 10 \tab 1 \tab - \tab 5 \tab 12 \tab 8 \tab 5 \tab 6 \tab 16 \tab 3 \tab 5 \tab 2 \tab 11 \tab 1 \tab - \tab 197 \tab 1 \tab 2 \tab 12 \tab 16 \tab - \tab 15 \tab 3 \tab 28 \tab 2 \tab 3 \tab 8 \tab 3 \tab 10 \tab 11 \tab 2 \tab 9 \tab 3 \tab - \tab 2 \tab - \tab 5 \tab 1 \tab 3 \tab 2 \tab 7 \tab 1 \tab 3 \tab 1 \tab 1 \tab 36 \tab 2 \tab 25 \tab 2 \tab 4 \tab 4 \tab 44 \tab - \tab 3 \tab 6 \tab 2 \tab - \tab 11 \tab 2 \tab 5 \tab - \tab 1 \tab - \tab 4 \tab 6 \tab 3 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 3 \tab 4 \tab 1 \tab 2 \tab 2 \tab 4 \tab 4 \tab 2 \tab 2 \tab 16 \tab - \tab 10 \tab 1 \tab 8 \tab 2 \tab 29 \tab 2 \tab 5 \tab 2 \tab 12 \tab 16 \tab 4 \tab 2 \tab 9 \tab 1 \tab 1 \tab 12 \tab 2 \tab 4 \tab 8 \tab 4 \tab 2 \tab 4 \tab 3 \tab 5 \tab 3 \tab 3 \tab - \tab 3 \tab - \tab 5 \tab 24 \tab 5 \tab 6 \tab 1 \tab 4 \tab 3 \tab 2 \tab 3 \tab 6 \tab 1 \tab 13 \tab - \tab 5 \tab 5 \tab 4 \tab 3 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 4 \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 10 \tab 2 \tab 7 \tab 19 \tab 2 \tab 5 \tab 25 \tab 3 \tab 9 \tab 170 \tab 10 \tab 2 \tab 15 \tab 2 \tab 20 \tab 25 \tab 6 \tab 18 \tab 10 \tab 8 \tab 20 \tab 11 \tab 3 \tab 8 \tab 14 \tab 20 \tab 8 \tab 8 \tab 1 \tab 10 \tab 222 \tab 13 \tab 176 \tab 12 \tab 2 \tab 13 \tab 9 \tab - \tab 4 \tab 1 \tab 3 \tab 22 \tab 10 \tab 1 \tab 58 \tab 3 \tab 20 \tab 20 \tab - \tab 6 \tab 6 \tab 2 \tab 5 \tab 2 \tab 15 \tab 8 \tab 4 \tab 4 \tab 15 \tab 14 \tab 11 \tab 20 \tab 29 \tab 3 \tab 7 \tab 2 \tab 19 \tab 8 \tab 6 \tab 508 \tab 6 \tab - \tab 29 \tab 33 \tab - \tab 52 \tab 4 \tab 47 \tab 4 \tab 6 \tab 10 \tab 1 \tab 16 \tab 13 \tab 1 \tab 17 \tab 4 \tab 1 \tab 5 \tab 2 \tab 13 \tab - \tab 4 \tab 4 \tab 12 \tab 1 \tab 1 \tab 2 \tab 3 \tab 47 \tab 9 \tab 33 \tab 3 \tab 3 \tab 10 \tab 41 \tab 2 \tab 2 \tab 10 \tab 2 \tab 2 \tab 25 \tab 1 \tab 9 \tab 7 \tab - \tab 4 \tab 6 \tab 7 \tab 1 \tab 5 \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 3 \tab 3 \tab 2 \tab 6 \tab 12 \tab 4 \tab 9 \tab 3 \tab 28 \tab 1 \tab 14 \tab 3 \tab 15 \tab 2 \tab 82 \tab 1 \tab 14 \tab 1 \tab 21 \tab 25 \tab 5 \tab 5 \tab 17 \tab 4 \tab 4 \tab 18 \tab 8 \tab 2 \tab 18 \tab 1 \tab 2 \tab 5 \tab 3 \tab 12 \tab 6 \tab 1 \tab 5 \tab 6 \tab - \tab 9 \tab 47 \tab 16 \tab 7 \tab 10 \tab 5 \tab 5 \tab 5 \tab 2 \tab 5 \tab 3 \tab 22 \tab - \tab 11 \tab 2 \tab 11 \tab 4 \tab - \tab - \tab 1 \tab 5 \tab - \tab 3 \tab 4 \tab 7 \tab 5 \tab 10 \tab 9 \tab 6 \tab 1 \tab 6 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 7 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 6 \tab 6 \tab 16 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 3 \tab 2 \tab - \tab 2 \tab 3 \tab 5 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 4 \tab - \tab - \tab 3 \tab 5 \tab 2 \tab - \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 10 \tab 6 \tab 2 \tab 14 \tab 2 \tab 9 \tab 21 \tab 2 \tab 5 \tab 152 \tab 10 \tab 1 \tab 14 \tab 4 \tab 25 \tab 27 \tab 3 \tab 13 \tab 8 \tab 6 \tab 12 \tab 12 \tab 1 \tab 7 \tab 18 \tab 17 \tab 5 \tab 6 \tab 4 \tab 3 \tab 261 \tab 3 \tab 142 \tab 10 \tab 1 \tab 7 \tab 16 \tab - \tab 4 \tab 2 \tab 7 \tab 34 \tab 17 \tab 2 \tab 59 \tab - \tab 10 \tab 16 \tab 2 \tab 5 \tab 6 \tab 1 \tab 6 \tab 2 \tab 11 \tab 5 \tab 4 \tab 13 \tab 12 \tab 11 \tab 6 \tab 16 \tab 36 \tab 4 \tab 4 \tab - \tab 20 \tab 4 \tab 2 \tab 502 \tab 3 \tab 1 \tab 17 \tab 26 \tab 3 \tab 40 \tab 3 \tab 46 \tab 5 \tab 6 \tab 8 \tab 4 \tab 12 \tab 10 \tab 1 \tab 13 \tab 4 \tab 1 \tab 6 \tab 2 \tab 5 \tab - \tab 7 \tab 5 \tab 9 \tab 8 \tab 3 \tab 2 \tab 2 \tab 61 \tab 3 \tab 53 \tab 9 \tab 5 \tab 5 \tab 63 \tab 1 \tab 3 \tab 7 \tab 3 \tab 5 \tab 16 \tab 5 \tab 11 \tab 2 \tab 1 \tab - \tab 4 \tab 9 \tab 4 \tab 8 \tab - \tab 2 \tab 7 \tab 2 \tab 4 \tab 2 \tab 1 \tab 5 \tab 4 \tab 2 \tab 9 \tab 7 \tab 1 \tab 4 \tab 7 \tab 25 \tab 2 \tab 11 \tab 3 \tab 9 \tab 5 \tab 59 \tab 1 \tab 18 \tab 2 \tab 25 \tab 18 \tab 4 \tab 2 \tab 21 \tab 2 \tab 5 \tab 21 \tab 3 \tab 2 \tab 6 \tab 2 \tab 4 \tab 4 \tab 1 \tab 10 \tab 6 \tab - \tab 3 \tab 4 \tab 1 \tab 5 \tab 54 \tab 13 \tab 14 \tab 9 \tab 2 \tab 5 \tab 7 \tab 6 \tab 6 \tab 4 \tab 16 \tab - \tab 10 \tab 9 \tab 9 \tab 9 \tab 1 \tab 1 \tab 1 \tab 13 \tab 6 \tab 1 \tab 3 \tab 5 \tab 6 \tab 5 \tab 8 \tab 1 \tab 2 \tab 6 \tab 3 \tab 7 \tab 1 \tab 5 \tab 3 \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab 3 \tab 6 \tab 2 \tab 14 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 2 \tab 5 \tab - \tab 1 \tab 3 \tab 6 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 4 \tab - \tab 2 \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 7 \tab 12 \tab 4 \tab 9 \tab 2 \tab 4 \tab 20 \tab - \tab 3 \tab 135 \tab 14 \tab 1 \tab 11 \tab 8 \tab 24 \tab 32 \tab 8 \tab 12 \tab 4 \tab 7 \tab 12 \tab 11 \tab 1 \tab 6 \tab 19 \tab 31 \tab 11 \tab 1 \tab 5 \tab 11 \tab 263 \tab 4 \tab 147 \tab 4 \tab 2 \tab 15 \tab 9 \tab - \tab 1 \tab 3 \tab 6 \tab 19 \tab 10 \tab 8 \tab 50 \tab 4 \tab 19 \tab 16 \tab 3 \tab 2 \tab 5 \tab 5 \tab 5 \tab 2 \tab 13 \tab 9 \tab 3 \tab 6 \tab 17 \tab 15 \tab 2 \tab 10 \tab 31 \tab 3 \tab 1 \tab - \tab 19 \tab 7 \tab 3 \tab 503 \tab 7 \tab 4 \tab 14 \tab 31 \tab 5 \tab 36 \tab 2 \tab 40 \tab 3 \tab 7 \tab 11 \tab 4 \tab 15 \tab 10 \tab 4 \tab 17 \tab 6 \tab 1 \tab 8 \tab 1 \tab 6 \tab - \tab 8 \tab 3 \tab 10 \tab 3 \tab 3 \tab 4 \tab 2 \tab 51 \tab 2 \tab 48 \tab 7 \tab - \tab 5 \tab 66 \tab - \tab 2 \tab 5 \tab 4 \tab 3 \tab 22 \tab 2 \tab 12 \tab 1 \tab 1 \tab - \tab 5 \tab 15 \tab 3 \tab 2 \tab - \tab 1 \tab 5 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab 6 \tab 4 \tab 4 \tab 16 \tab 5 \tab 7 \tab 3 \tab 26 \tab 1 \tab 11 \tab - \tab 8 \tab 1 \tab 53 \tab 1 \tab 16 \tab 2 \tab 19 \tab 39 \tab 1 \tab - \tab 14 \tab 5 \tab 8 \tab 10 \tab 3 \tab 2 \tab 13 \tab 2 \tab 2 \tab 7 \tab 2 \tab 12 \tab 2 \tab 1 \tab 2 \tab 4 \tab 1 \tab 8 \tab 34 \tab 16 \tab 17 \tab 11 \tab 3 \tab 3 \tab 6 \tab 3 \tab 6 \tab 5 \tab 18 \tab - \tab 10 \tab 1 \tab 8 \tab 4 \tab 1 \tab 2 \tab 1 \tab 6 \tab 2 \tab 1 \tab - \tab 6 \tab 2 \tab 6 \tab 11 \tab 7 \tab 1 \tab 6 \tab - \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 7 \tab 2 \tab 1 \tab 4 \tab 2 \tab 3 \tab 3 \tab 4 \tab 12 \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 4 \tab 2 \tab 4 \tab - \tab 2 \tab 5 \tab 5 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 4 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 9 \tab 7 \tab 3 \tab 10 \tab - \tab 14 \tab 19 \tab 5 \tab 7 \tab 171 \tab 13 \tab 2 \tab 11 \tab 4 \tab 28 \tab 33 \tab 9 \tab 6 \tab 5 \tab 12 \tab 10 \tab 10 \tab 3 \tab 5 \tab 21 \tab 20 \tab 6 \tab 3 \tab 4 \tab 5 \tab 248 \tab 6 \tab 139 \tab 14 \tab 1 \tab 13 \tab 8 \tab - \tab 4 \tab 5 \tab 9 \tab 28 \tab 4 \tab 2 \tab 43 \tab - \tab 18 \tab 20 \tab 3 \tab 2 \tab 7 \tab - \tab 5 \tab - \tab 5 \tab 7 \tab 2 \tab 10 \tab 10 \tab 16 \tab 6 \tab 12 \tab 24 \tab 8 \tab 4 \tab - \tab 18 \tab 9 \tab 2 \tab 518 \tab 5 \tab 3 \tab 20 \tab 28 \tab 6 \tab 54 \tab 4 \tab 44 \tab 1 \tab 2 \tab 6 \tab 1 \tab 11 \tab 6 \tab 6 \tab 12 \tab 9 \tab 1 \tab 1 \tab - \tab 7 \tab 2 \tab 3 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab - \tab 61 \tab 4 \tab 51 \tab 9 \tab 1 \tab 8 \tab 60 \tab 4 \tab 3 \tab 9 \tab 1 \tab 2 \tab 20 \tab 1 \tab 4 \tab 3 \tab 3 \tab - \tab 4 \tab 9 \tab 7 \tab 8 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 4 \tab 3 \tab 9 \tab 7 \tab 11 \tab 3 \tab 5 \tab 1 \tab 23 \tab - \tab 13 \tab 3 \tab 18 \tab 4 \tab 48 \tab 2 \tab 19 \tab 2 \tab 23 \tab 19 \tab 4 \tab 4 \tab 11 \tab 5 \tab 11 \tab 23 \tab 2 \tab 4 \tab 11 \tab 2 \tab - \tab 6 \tab 1 \tab 6 \tab 5 \tab 2 \tab 5 \tab 5 \tab 2 \tab 7 \tab 45 \tab 11 \tab 7 \tab 11 \tab - \tab 11 \tab 5 \tab 4 \tab 4 \tab 1 \tab 11 \tab 1 \tab 8 \tab 2 \tab 4 \tab 7 \tab - \tab - \tab 1 \tab 8 \tab 1 \tab 3 \tab 5 \tab 9 \tab 3 \tab 6 \tab 9 \tab 2 \tab - \tab 4 \tab - \tab 6 \tab - \tab 3 \tab 2 \tab 5 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 5 \tab 13 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 4 \tab 2 \tab - \tab 1 \tab 5 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 6 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 10 \tab 4 \tab 9 \tab 10 \tab 3 \tab 5 \tab 20 \tab 3 \tab 4 \tab 131 \tab 9 \tab 1 \tab 14 \tab 4 \tab 15 \tab 44 \tab 8 \tab 9 \tab 5 \tab 11 \tab 17 \tab 8 \tab 1 \tab 4 \tab 16 \tab 25 \tab 5 \tab 4 \tab 4 \tab 10 \tab 249 \tab - \tab 127 \tab 5 \tab 1 \tab 11 \tab 8 \tab 1 \tab 5 \tab 2 \tab 3 \tab 28 \tab 7 \tab 2 \tab 55 \tab 5 \tab 17 \tab 29 \tab 4 \tab 2 \tab 4 \tab 1 \tab 4 \tab - \tab 7 \tab 3 \tab 2 \tab 6 \tab 16 \tab 16 \tab 5 \tab 11 \tab 25 \tab 1 \tab 3 \tab 1 \tab 21 \tab 6 \tab - \tab 526 \tab 4 \tab 3 \tab 25 \tab 19 \tab 2 \tab 50 \tab - \tab 53 \tab 2 \tab 8 \tab 7 \tab 4 \tab 10 \tab 11 \tab 2 \tab 9 \tab 7 \tab 3 \tab 4 \tab 2 \tab 9 \tab 1 \tab 3 \tab 2 \tab 7 \tab 4 \tab - \tab 4 \tab 8 \tab 53 \tab 5 \tab 41 \tab 6 \tab - \tab 10 \tab 63 \tab 1 \tab 2 \tab 4 \tab 1 \tab 3 \tab 16 \tab 2 \tab 5 \tab 3 \tab 4 \tab 1 \tab 5 \tab 8 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 4 \tab 2 \tab 6 \tab 8 \tab 13 \tab 3 \tab 7 \tab 7 \tab 31 \tab - \tab 14 \tab 2 \tab 21 \tab 2 \tab 44 \tab 1 \tab 15 \tab 2 \tab 16 \tab 28 \tab 4 \tab 2 \tab 11 \tab 3 \tab 9 \tab 13 \tab 7 \tab 3 \tab 11 \tab 1 \tab - \tab 5 \tab 1 \tab 4 \tab 7 \tab - \tab 2 \tab 6 \tab 4 \tab 9 \tab 57 \tab 5 \tab 16 \tab 13 \tab - \tab 11 \tab 7 \tab 1 \tab 5 \tab 1 \tab 24 \tab 1 \tab 8 \tab 3 \tab 1 \tab 12 \tab - \tab 3 \tab 1 \tab 6 \tab 1 \tab - \tab - \tab 4 \tab 4 \tab 6 \tab 6 \tab 1 \tab - \tab 3 \tab 2 \tab 5 \tab 2 \tab 5 \tab - \tab 6 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 4 \tab - \tab 8 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 5 \tab 4 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 9 \tab 3 \tab 10 \tab 7 \tab 2 \tab 9 \tab 21 \tab 3 \tab 4 \tab 110 \tab 12 \tab 3 \tab 16 \tab 5 \tab 31 \tab 32 \tab 5 \tab 15 \tab 3 \tab 7 \tab 15 \tab 10 \tab - \tab 5 \tab 23 \tab 37 \tab 5 \tab 5 \tab - \tab 8 \tab 268 \tab 7 \tab 131 \tab 6 \tab 2 \tab 7 \tab 8 \tab - \tab 4 \tab 2 \tab 5 \tab 21 \tab 9 \tab 5 \tab 78 \tab 4 \tab 18 \tab 10 \tab 2 \tab 3 \tab 7 \tab 2 \tab 4 \tab 1 \tab 10 \tab 6 \tab 5 \tab 6 \tab 16 \tab 13 \tab 8 \tab 5 \tab 24 \tab 7 \tab - \tab 1 \tab 25 \tab 4 \tab 7 \tab 514 \tab 1 \tab 2 \tab 22 \tab 26 \tab 3 \tab 39 \tab 2 \tab 38 \tab - \tab 4 \tab 7 \tab - \tab 13 \tab 18 \tab 1 \tab 14 \tab 6 \tab 1 \tab 2 \tab 2 \tab 8 \tab - \tab 10 \tab 4 \tab 7 \tab 1 \tab 3 \tab 4 \tab 4 \tab 51 \tab 2 \tab 39 \tab 6 \tab 2 \tab 10 \tab 86 \tab 5 \tab - \tab 3 \tab 4 \tab 3 \tab 33 \tab 1 \tab 6 \tab 3 \tab 1 \tab - \tab 3 \tab 11 \tab 2 \tab 6 \tab - \tab 3 \tab 8 \tab 2 \tab 1 \tab 2 \tab - \tab 4 \tab 5 \tab 3 \tab 6 \tab 7 \tab 3 \tab 5 \tab 3 \tab 24 \tab 4 \tab 16 \tab 1 \tab 19 \tab 1 \tab 30 \tab 2 \tab 16 \tab - \tab 16 \tab 22 \tab 6 \tab 4 \tab 16 \tab 7 \tab 3 \tab 15 \tab 5 \tab 1 \tab 10 \tab 2 \tab 1 \tab 6 \tab 1 \tab 13 \tab 5 \tab - \tab - \tab 3 \tab 4 \tab 3 \tab 46 \tab 12 \tab 14 \tab 15 \tab - \tab 10 \tab 12 \tab 1 \tab 6 \tab 2 \tab 19 \tab - \tab 13 \tab 7 \tab 4 \tab 8 \tab - \tab 1 \tab 2 \tab 5 \tab 4 \tab - \tab - \tab 6 \tab 2 \tab 9 \tab 10 \tab 3 \tab 2 \tab - \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 6 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 8 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 3 \tab 6 \tab 6 \tab 6 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 5 \tab 4 \tab 11 \tab 9 \tab 2 \tab 4 \tab 9 \tab - \tab 3 \tab 109 \tab 11 \tab 1 \tab 18 \tab 11 \tab 25 \tab 40 \tab 6 \tab 14 \tab 5 \tab 10 \tab 16 \tab 16 \tab - \tab 8 \tab 23 \tab 26 \tab 7 \tab 3 \tab 2 \tab 7 \tab 306 \tab 9 \tab 164 \tab 16 \tab 4 \tab 8 \tab 5 \tab - \tab 11 \tab - \tab 7 \tab 35 \tab 13 \tab 3 \tab 67 \tab 1 \tab 23 \tab 10 \tab 3 \tab 3 \tab 7 \tab 1 \tab 3 \tab 2 \tab 6 \tab 5 \tab 2 \tab 12 \tab 29 \tab 14 \tab 6 \tab 12 \tab 28 \tab 4 \tab 2 \tab 2 \tab 32 \tab 2 \tab 5 \tab 2049 \tab 3 \tab - \tab 20 \tab 35 \tab 4 \tab 40 \tab 4 \tab 49 \tab 3 \tab 6 \tab 11 \tab 1 \tab 16 \tab 25 \tab 3 \tab 16 \tab 3 \tab 1 \tab 6 \tab 1 \tab 9 \tab - \tab 4 \tab 2 \tab 6 \tab 1 \tab 5 \tab 3 \tab 6 \tab 53 \tab 6 \tab 36 \tab 4 \tab 3 \tab 8 \tab 55 \tab 4 \tab 2 \tab 3 \tab 4 \tab 1 \tab 41 \tab - \tab 7 \tab 1 \tab - \tab - \tab 5 \tab 14 \tab 2 \tab 9 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 3 \tab 7 \tab 9 \tab 12 \tab 5 \tab 7 \tab 6 \tab 33 \tab 2 \tab 7 \tab 6 \tab 17 \tab 2 \tab 37 \tab 3 \tab 13 \tab - \tab 23 \tab 27 \tab 3 \tab 2 \tab 11 \tab 9 \tab 7 \tab 19 \tab 4 \tab 1 \tab 13 \tab 2 \tab 2 \tab 8 \tab 4 \tab 6 \tab 3 \tab - \tab 1 \tab 9 \tab 1 \tab 4 \tab 59 \tab 16 \tab 11 \tab 5 \tab - \tab 8 \tab 9 \tab 7 \tab 6 \tab 2 \tab 24 \tab 1 \tab 10 \tab 3 \tab 1 \tab 8 \tab - \tab 2 \tab 1 \tab 7 \tab 2 \tab 2 \tab 4 \tab 4 \tab 1 \tab 4 \tab 8 \tab 3 \tab - \tab 5 \tab 4 \tab 1 \tab - \tab 2 \tab - \tab 7 \tab - \tab 6 \tab 2 \tab - \tab - \tab 2 \tab - \tab 6 \tab 2 \tab 4 \tab 3 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab 8 \tab 8 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 3 \tab 2 \tab - \tab 6 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab 2 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 6 \tab 6 \tab 2 \tab 7 \tab 3 \tab 9 \tab 18 \tab 2 \tab 4 \tab 71 \tab 8 \tab 1 \tab 17 \tab 5 \tab 21 \tab 25 \tab 6 \tab 9 \tab 4 \tab 6 \tab 8 \tab 11 \tab - \tab 4 \tab 14 \tab 13 \tab 4 \tab 5 \tab 2 \tab 7 \tab 184 \tab 4 \tab 91 \tab 5 \tab 1 \tab 6 \tab 6 \tab - \tab - \tab 2 \tab 3 \tab 13 \tab 5 \tab 1 \tab 35 \tab - \tab 19 \tab 6 \tab 1 \tab - \tab 7 \tab - \tab 5 \tab 4 \tab 9 \tab 3 \tab 4 \tab 5 \tab 8 \tab 12 \tab 10 \tab 8 \tab 23 \tab - \tab 2 \tab 2 \tab 14 \tab 3 \tab 2 \tab 354 \tab 3 \tab 2 \tab 13 \tab 28 \tab - \tab 27 \tab - \tab 38 \tab 3 \tab 14 \tab 11 \tab - \tab 9 \tab 9 \tab 3 \tab 10 \tab 4 \tab 2 \tab 5 \tab 1 \tab 6 \tab 1 \tab 5 \tab 1 \tab 7 \tab 3 \tab 2 \tab 2 \tab 1 \tab 42 \tab 3 \tab 30 \tab 10 \tab 2 \tab 4 \tab 44 \tab 5 \tab 3 \tab 2 \tab 1 \tab 2 \tab 12 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 6 \tab 2 \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 3 \tab 2 \tab 4 \tab 6 \tab 1 \tab 2 \tab 5 \tab 16 \tab 1 \tab 9 \tab 2 \tab 11 \tab 1 \tab 28 \tab 3 \tab 17 \tab 1 \tab 12 \tab 27 \tab 4 \tab 1 \tab 14 \tab 4 \tab 8 \tab 11 \tab 4 \tab 3 \tab 1 \tab 2 \tab 4 \tab 3 \tab - \tab 6 \tab 3 \tab 1 \tab 3 \tab 2 \tab 2 \tab 6 \tab 34 \tab 12 \tab 10 \tab 8 \tab - \tab 7 \tab 7 \tab 4 \tab 6 \tab - \tab 22 \tab 2 \tab 11 \tab - \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab - \tab 4 \tab 2 \tab 4 \tab 2 \tab 4 \tab 9 \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 4 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab 6 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 4 \tab 4 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2 \tab 4 \tab 4 \tab 10 \tab 2 \tab 9 \tab 15 \tab - \tab 3 \tab 81 \tab 9 \tab 3 \tab 11 \tab 4 \tab 30 \tab 32 \tab 6 \tab 11 \tab 5 \tab 4 \tab 12 \tab 10 \tab - \tab 2 \tab 9 \tab 34 \tab 1 \tab 4 \tab 2 \tab 9 \tab 185 \tab 2 \tab 79 \tab 10 \tab - \tab 5 \tab 6 \tab - \tab 1 \tab 3 \tab 6 \tab 24 \tab 5 \tab 6 \tab 35 \tab 3 \tab 8 \tab 11 \tab 2 \tab 1 \tab 8 \tab 1 \tab 1 \tab - \tab 6 \tab 3 \tab 1 \tab 8 \tab 17 \tab 8 \tab 6 \tab 14 \tab 15 \tab 4 \tab 2 \tab 2 \tab 18 \tab 1 \tab 1 \tab 354 \tab 4 \tab 1 \tab 12 \tab 15 \tab 3 \tab 23 \tab 2 \tab 25 \tab 2 \tab 1 \tab 9 \tab 1 \tab 6 \tab 8 \tab 2 \tab 15 \tab 1 \tab 2 \tab 4 \tab 1 \tab 5 \tab 1 \tab 6 \tab 1 \tab 4 \tab 3 \tab 3 \tab 1 \tab 4 \tab 43 \tab 3 \tab 33 \tab 4 \tab 1 \tab 9 \tab 38 \tab 6 \tab 1 \tab 3 \tab 3 \tab 2 \tab 11 \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 7 \tab 12 \tab - \tab 8 \tab 1 \tab 2 \tab 3 \tab - \tab 3 \tab - \tab - \tab 4 \tab 5 \tab 5 \tab 4 \tab 6 \tab 3 \tab 3 \tab 5 \tab 13 \tab - \tab 9 \tab 2 \tab 15 \tab 1 \tab 32 \tab 2 \tab 21 \tab - \tab 8 \tab 27 \tab 1 \tab 2 \tab 18 \tab 4 \tab 5 \tab 16 \tab 10 \tab - \tab 6 \tab 4 \tab 3 \tab 1 \tab 2 \tab 5 \tab 7 \tab - \tab 1 \tab 3 \tab 2 \tab 7 \tab 39 \tab 7 \tab 6 \tab 3 \tab - \tab 5 \tab 6 \tab 3 \tab 3 \tab - \tab 14 \tab - \tab 12 \tab 1 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 8 \tab 2 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 2 \tab 3 \tab 3 \tab 1 \tab 3 \tab 5 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 2 \tab 9 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 3 \tab - \tab 1 \tab 8 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 5 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 4 \tab 4 \tab 4 \tab 4 \tab 1 \tab 4 \tab 22 \tab 3 \tab 2 \tab 48 \tab 12 \tab 1 \tab 10 \tab 6 \tab 16 \tab 25 \tab 4 \tab 11 \tab 3 \tab 12 \tab 5 \tab 2 \tab - \tab 4 \tab 7 \tab 29 \tab 1 \tab 5 \tab 1 \tab 3 \tab 200 \tab 3 \tab 119 \tab 4 \tab 1 \tab 3 \tab 8 \tab - \tab 3 \tab 2 \tab 3 \tab 13 \tab 7 \tab 4 \tab 43 \tab 4 \tab 22 \tab 3 \tab - \tab 5 \tab - \tab - \tab 4 \tab - \tab 6 \tab 2 \tab 3 \tab 7 \tab 5 \tab 7 \tab 5 \tab 12 \tab 19 \tab 1 \tab 2 \tab - \tab 22 \tab 2 \tab 4 \tab 341 \tab 5 \tab 3 \tab 16 \tab 22 \tab 4 \tab 27 \tab 2 \tab 35 \tab 1 \tab 4 \tab 6 \tab 2 \tab 4 \tab 2 \tab 3 \tab 12 \tab 1 \tab 5 \tab 4 \tab 1 \tab 7 \tab 1 \tab 7 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 41 \tab 5 \tab 6 \tab 7 \tab 2 \tab 7 \tab 42 \tab 1 \tab - \tab 3 \tab 3 \tab 2 \tab 31 \tab 5 \tab 4 \tab 2 \tab - \tab - \tab 3 \tab 10 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 5 \tab - \tab 3 \tab 3 \tab 6 \tab 1 \tab 3 \tab 6 \tab 19 \tab 1 \tab 7 \tab 3 \tab 9 \tab 4 \tab 28 \tab 2 \tab 13 \tab 2 \tab 16 \tab 17 \tab 3 \tab 2 \tab 4 \tab 2 \tab 3 \tab 9 \tab 3 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 4 \tab 3 \tab - \tab 1 \tab 3 \tab - \tab 4 \tab 29 \tab - \tab - \tab - \tab - \tab 7 \tab 3 \tab 3 \tab 2 \tab 1 \tab 5 \tab - \tab 9 \tab 1 \tab 3 \tab 5 \tab - \tab - \tab - \tab 4 \tab - \tab 3 \tab 3 \tab 6 \tab 1 \tab 3 \tab 7 \tab 3 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 4 \tab 2 \tab 1 \tab 2 \tab 3 \tab 10 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 75 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 4 \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab 7 \tab 3 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 5 \tab 3 \tab 3 \tab 6 \tab 2 \tab 8 \tab 18 \tab 5 \tab 7 \tab 75 \tab 10 \tab 1 \tab 13 \tab 6 \tab 28 \tab 38 \tab 11 \tab 15 \tab 3 \tab 9 \tab 13 \tab 9 \tab 1 \tab 2 \tab 15 \tab 27 \tab 10 \tab 4 \tab 7 \tab 6 \tab 242 \tab 4 \tab 125 \tab 7 \tab 1 \tab 8 \tab 7 \tab - \tab 2 \tab 4 \tab 4 \tab 22 \tab 12 \tab 4 \tab 42 \tab 2 \tab 14 \tab 4 \tab 2 \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab 3 \tab 4 \tab 1 \tab 7 \tab 13 \tab 9 \tab 5 \tab 10 \tab 27 \tab 2 \tab 3 \tab 2 \tab 18 \tab 5 \tab 4 \tab 459 \tab 11 \tab 4 \tab 14 \tab 27 \tab 7 \tab 31 \tab 3 \tab 49 \tab - \tab 6 \tab 14 \tab 4 \tab 13 \tab 7 \tab 2 \tab 9 \tab 5 \tab 7 \tab 3 \tab 1 \tab 12 \tab - \tab 8 \tab 1 \tab 5 \tab 3 \tab 1 \tab 1 \tab 2 \tab 66 \tab 5 \tab 28 \tab 5 \tab 1 \tab 3 \tab 67 \tab 4 \tab 4 \tab 3 \tab 3 \tab 1 \tab 14 \tab 1 \tab 15 \tab 3 \tab - \tab - \tab 3 \tab 14 \tab - \tab 5 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 5 \tab 3 \tab 11 \tab 2 \tab 2 \tab 4 \tab 19 \tab 4 \tab 20 \tab 3 \tab 9 \tab 1 \tab 26 \tab 1 \tab 15 \tab - \tab 20 \tab 31 \tab 4 \tab 1 \tab 7 \tab 4 \tab 5 \tab 14 \tab 8 \tab 3 \tab 3 \tab 3 \tab 3 \tab 7 \tab 1 \tab 15 \tab 8 \tab 1 \tab 2 \tab 4 \tab 1 \tab 4 \tab 53 \tab 15 \tab 16 \tab 13 \tab - \tab 5 \tab 8 \tab 4 \tab 3 \tab 1 \tab 14 \tab 3 \tab 10 \tab 8 \tab 5 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 7 \tab 2 \tab 5 \tab 5 \tab 7 \tab 5 \tab 1 \tab 3 \tab 2 \tab - \tab 4 \tab - \tab 5 \tab - \tab 3 \tab - \tab 4 \tab 4 \tab - \tab 2 \tab 2 \tab 3 \tab 2 \tab 7 \tab 9 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 4 \tab 5 \tab 3 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 4 \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 9 \tab 8 \tab 3 \tab 11 \tab - \tab 10 \tab 9 \tab 1 \tab 8 \tab 93 \tab 10 \tab - \tab 13 \tab 8 \tab 32 \tab 37 \tab 11 \tab 19 \tab 8 \tab 10 \tab 13 \tab 8 \tab 1 \tab 4 \tab 4 \tab 27 \tab 11 \tab 4 \tab 4 \tab 4 \tab 266 \tab 7 \tab 99 \tab 7 \tab 2 \tab 11 \tab 5 \tab - \tab 3 \tab 4 \tab 4 \tab 36 \tab 5 \tab 8 \tab 66 \tab 2 \tab 27 \tab 3 \tab 5 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab 5 \tab 1 \tab 3 \tab 6 \tab 15 \tab 12 \tab 11 \tab 13 \tab 22 \tab 2 \tab 2 \tab 2 \tab 32 \tab 4 \tab 2 \tab 548 \tab 6 \tab 3 \tab 20 \tab 21 \tab 6 \tab 37 \tab - \tab 51 \tab - \tab 11 \tab 15 \tab - \tab 11 \tab 3 \tab 5 \tab 16 \tab 5 \tab 7 \tab 6 \tab 1 \tab 6 \tab 1 \tab 9 \tab 4 \tab 7 \tab - \tab - \tab 2 \tab 3 \tab 68 \tab 3 \tab 53 \tab 8 \tab 2 \tab 4 \tab 71 \tab 3 \tab 4 \tab 7 \tab 2 \tab 1 \tab 37 \tab 6 \tab 14 \tab 4 \tab - \tab - \tab 3 \tab 10 \tab 1 \tab 5 \tab - \tab 3 \tab 4 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 1 \tab 8 \tab 5 \tab 10 \tab 6 \tab 9 \tab 3 \tab 18 \tab 6 \tab 21 \tab - \tab 7 \tab 4 \tab 37 \tab 2 \tab 21 \tab 2 \tab 17 \tab 33 \tab 6 \tab 2 \tab 8 \tab 6 \tab 10 \tab 22 \tab 6 \tab 2 \tab 13 \tab 2 \tab 2 \tab 5 \tab 1 \tab 13 \tab 6 \tab - \tab - \tab 5 \tab - \tab 9 \tab 56 \tab 15 \tab 19 \tab 5 \tab - \tab 4 \tab 8 \tab 5 \tab 10 \tab 1 \tab 19 \tab 4 \tab 8 \tab 5 \tab 4 \tab 2 \tab - \tab - \tab 1 \tab 5 \tab 3 \tab 3 \tab - \tab 7 \tab 2 \tab 3 \tab 4 \tab 5 \tab 2 \tab 3 \tab - \tab 1 \tab 3 \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 3 \tab 2 \tab 3 \tab 8 \tab 11 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 7 \tab - \tab 1 \tab 4 \tab 3 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 5 \tab - \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 17 \tab 1 \tab 3 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 9 \tab 7 \tab 7 \tab 9 \tab 4 \tab 8 \tab 16 \tab 2 \tab 4 \tab 44 \tab 7 \tab 1 \tab 13 \tab 6 \tab 30 \tab 39 \tab 4 \tab 12 \tab 3 \tab 5 \tab 7 \tab 12 \tab 1 \tab 3 \tab 5 \tab 20 \tab 8 \tab 3 \tab 2 \tab 5 \tab 208 \tab 3 \tab 115 \tab 1 \tab 2 \tab 6 \tab 8 \tab - \tab 2 \tab 3 \tab 9 \tab 14 \tab 7 \tab 5 \tab 45 \tab 1 \tab 4 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 3 \tab 4 \tab 8 \tab 6 \tab 14 \tab 4 \tab 14 \tab 18 \tab 4 \tab 2 \tab 2 \tab 20 \tab 4 \tab 6 \tab 464 \tab 2 \tab 1 \tab 12 \tab 26 \tab 2 \tab 32 \tab 2 \tab 46 \tab 2 \tab 9 \tab 14 \tab - \tab 8 \tab 7 \tab 5 \tab 14 \tab 3 \tab 10 \tab 3 \tab - \tab 8 \tab 3 \tab 4 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 25 \tab 3 \tab 16 \tab 3 \tab 3 \tab 5 \tab 38 \tab 3 \tab 2 \tab 3 \tab - \tab - \tab 12 \tab 2 \tab 20 \tab 2 \tab - \tab 1 \tab 6 \tab 10 \tab 3 \tab 9 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 6 \tab 6 \tab 4 \tab 6 \tab 4 \tab 3 \tab 26 \tab - \tab 4 \tab - \tab 4 \tab 4 \tab 31 \tab 2 \tab 25 \tab 2 \tab 10 \tab 23 \tab 4 \tab 1 \tab 13 \tab 6 \tab 9 \tab 22 \tab 3 \tab 7 \tab 6 \tab 2 \tab 1 \tab 3 \tab 1 \tab 7 \tab 3 \tab 2 \tab 1 \tab 6 \tab 1 \tab 10 \tab 51 \tab 20 \tab 28 \tab 16 \tab - \tab 12 \tab 9 \tab 2 \tab 3 \tab 5 \tab 16 \tab - \tab 9 \tab 5 \tab 5 \tab 3 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 3 \tab 5 \tab 6 \tab 3 \tab 4 \tab 7 \tab 7 \tab 3 \tab 4 \tab 1 \tab 1 \tab - \tab 6 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 7 \tab 4 \tab 16 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 7 \tab 10 \tab 3 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab 6 \tab 1 \tab 1 \tab 7 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2348 \cr
   2021 \tab 13 \tab 6 \tab 12 \tab 18 \tab 4 \tab 10 \tab 33 \tab 2 \tab 4 \tab 86 \tab 12 \tab 4 \tab 30 \tab 10 \tab 41 \tab 55 \tab 9 \tab 36 \tab 12 \tab 9 \tab 11 \tab 18 \tab - \tab 8 \tab 11 \tab 59 \tab 16 \tab 8 \tab 4 \tab 8 \tab 436 \tab 15 \tab 96 \tab 7 \tab 4 \tab 10 \tab 10 \tab - \tab 15 \tab 6 \tab 15 \tab 35 \tab 11 \tab 10 \tab 83 \tab 6 \tab 25 \tab 13 \tab 3 \tab 2 \tab 6 \tab 1 \tab 7 \tab 1 \tab 7 \tab 4 \tab 5 \tab 7 \tab 27 \tab 13 \tab 7 \tab 27 \tab 46 \tab 3 \tab 2 \tab 5 \tab 47 \tab 5 \tab 4 \tab 563 \tab 12 \tab 6 \tab 19 \tab 45 \tab 4 \tab 39 \tab 7 \tab 75 \tab 2 \tab 5 \tab 22 \tab 2 \tab 23 \tab 11 \tab 5 \tab 16 \tab 9 \tab 7 \tab 7 \tab 4 \tab 12 \tab 3 \tab 13 \tab 5 \tab 15 \tab 5 \tab 5 \tab 7 \tab 4 \tab 101 \tab 4 \tab 38 \tab 12 \tab 6 \tab 20 \tab 144 \tab 5 \tab 3 \tab 8 \tab 6 \tab 11 \tab 17 \tab 2 \tab 8 \tab 2 \tab 6 \tab 1 \tab 13 \tab 14 \tab 4 \tab 10 \tab - \tab 3 \tab 1 \tab 2 \tab 6 \tab 4 \tab - \tab 7 \tab 6 \tab 7 \tab 8 \tab 16 \tab 5 \tab 6 \tab 12 \tab 40 \tab 3 \tab 15 \tab 2 \tab 14 \tab 4 \tab 27 \tab 5 \tab 30 \tab 4 \tab 28 \tab 51 \tab 2 \tab 5 \tab 12 \tab 13 \tab 11 \tab 28 \tab 19 \tab 4 \tab 12 \tab 3 \tab 6 \tab 12 \tab 2 \tab 18 \tab 13 \tab 1 \tab 1 \tab 7 \tab 4 \tab 12 \tab 81 \tab - \tab - \tab - \tab - \tab 8 \tab 7 \tab 6 \tab 8 \tab 3 \tab 24 \tab 81 \tab 15 \tab 14 \tab 8 \tab 2 \tab - \tab 2 \tab 4 \tab 9 \tab 6 \tab 6 \tab 8 \tab 12 \tab 3 \tab 15 \tab 15 \tab 13 \tab 5 \tab 3 \tab 1 \tab 5 \tab 7 \tab 9 \tab 2 \tab 5 \tab 2 \tab 1 \tab 6 \tab 1 \tab 4 \tab 1 \tab 3 \tab 5 \tab 4 \tab 29 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 8 \tab - \tab - \tab - \tab 2 \tab 5 \tab 3 \tab 1 \tab 2 \tab 3 \tab - \tab 2 \tab 2 \tab 2 \tab 3 \tab 5 \tab 5 \tab 4 \tab - \tab 9 \tab - \tab 1 \tab 26 \tab 5 \tab 1 \tab - \tab 3 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 11 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 4032 \cr
   2022 \tab 12 \tab 4 \tab 13 \tab 13 \tab 1 \tab 7 \tab 18 \tab 2 \tab 3 \tab 64 \tab 22 \tab 2 \tab 31 \tab 11 \tab 25 \tab 63 \tab 7 \tab 32 \tab 6 \tab 8 \tab 13 \tab 19 \tab 1 \tab 3 \tab 4 \tab 41 \tab 10 \tab 6 \tab 4 \tab 3 \tab 394 \tab 11 \tab 106 \tab 7 \tab - \tab 11 \tab 11 \tab - \tab - \tab 1 \tab 6 \tab 25 \tab 15 \tab 8 \tab 87 \tab - \tab 29 \tab 5 \tab - \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 4 \tab 2 \tab 4 \tab 10 \tab 17 \tab 9 \tab 4 \tab 21 \tab 32 \tab 5 \tab 2 \tab 3 \tab 44 \tab 4 \tab 5 \tab 623 \tab 8 \tab 8 \tab 22 \tab 34 \tab 8 \tab 42 \tab 3 \tab 78 \tab 1 \tab 6 \tab 28 \tab 1 \tab 18 \tab 7 \tab 8 \tab 15 \tab 4 \tab 5 \tab 7 \tab 3 \tab 9 \tab 2 \tab 7 \tab 4 \tab 4 \tab 3 \tab 2 \tab 4 \tab 4 \tab 64 \tab 8 \tab 42 \tab 5 \tab 3 \tab 8 \tab 157 \tab 3 \tab - \tab 6 \tab 3 \tab 2 \tab 15 \tab 5 \tab 7 \tab 4 \tab 2 \tab 1 \tab 4 \tab 17 \tab 4 \tab 9 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 4 \tab 3 \tab 11 \tab 3 \tab 11 \tab 4 \tab 10 \tab 5 \tab 32 \tab 3 \tab 18 \tab 1 \tab 14 \tab 4 \tab 27 \tab 2 \tab 22 \tab 2 \tab 19 \tab 48 \tab 3 \tab 5 \tab 9 \tab 9 \tab 13 \tab 26 \tab 10 \tab 2 \tab 9 \tab 3 \tab - \tab 6 \tab 1 \tab 4 \tab 4 \tab 3 \tab 1 \tab 9 \tab 3 \tab 13 \tab 82 \tab - \tab - \tab - \tab - \tab 15 \tab 16 \tab 6 \tab 7 \tab 2 \tab 18 \tab 77 \tab 11 \tab 5 \tab 7 \tab 5 \tab - \tab 2 \tab 3 \tab 3 \tab 8 \tab 2 \tab 6 \tab 12 \tab 6 \tab 6 \tab 4 \tab 3 \tab 3 \tab 4 \tab 1 \tab 4 \tab 1 \tab 6 \tab 1 \tab 8 \tab - \tab 1 \tab 6 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 3 \tab 15 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 8 \tab - \tab - \tab 2 \tab 3 \tab 4 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 6 \tab 5 \tab 5 \tab 2 \tab 5 \tab - \tab 1 \tab 29 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 10 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 157 \tab 114 \tab 124 \tab 227 \tab 48 \tab 166 \tab 466 \tab 60 \tab 137 \tab 3438 \tab 294 \tab 50 \tab 331 \tab 140 \tab 616 \tab 800 \tab 152 \tab 336 \tab 197 \tab 213 \tab 310 \tab 299 \tab 23 \tab 138 \tab 784 \tab 577 \tab 176 \tab 116 \tab 87 \tab 132 \tab 6503 \tab 149 \tab 4164 \tab 217 \tab 47 \tab 229 \tab 221 \tab 14 \tab 107 \tab 63 \tab 189 \tab 565 \tab 230 \tab 95 \tab 1193 \tab 68 \tab 394 \tab 312 \tab 68 \tab 54 \tab 101 \tab 39 \tab 119 \tab 43 \tab 238 \tab 117 \tab 73 \tab 195 \tab 378 \tab 307 \tab 191 \tab 395 \tab 748 \tab 107 \tab 76 \tab 31 \tab 514 \tab 113 \tab 79 \tab 12803 \tab 122 \tab 55 \tab 440 \tab 670 \tab 65 \tab 1158 \tab 75 \tab 1092 \tab 60 \tab 140 \tab 234 \tab 66 \tab 284 \tab 258 \tab 86 \tab 376 \tab 111 \tab 68 \tab 237 \tab 36 \tab 197 \tab 25 \tab 179 \tab 114 \tab 192 \tab 66 \tab 65 \tab 62 \tab 92 \tab 1363 \tab 89 \tab 907 \tab 156 \tab 76 \tab 217 \tab 1636 \tab 74 \tab 78 \tab 131 \tab 63 \tab 63 \tab 443 \tab 60 \tab 200 \tab 94 \tab 30 \tab 19 \tab 121 \tab 277 \tab 78 \tab 149 \tab 10 \tab 53 \tab 79 \tab 38 \tab 38 \tab 45 \tab 40 \tab 95 \tab 79 \tab 147 \tab 146 \tab 217 \tab 85 \tab 137 \tab 108 \tab 655 \tab 61 \tab 299 \tab 83 \tab 313 \tab 69 \tab 1359 \tab 64 \tab 446 \tab 62 \tab 471 \tab 610 \tab 115 \tab 61 \tab 423 \tab 120 \tab 160 \tab 389 \tab 130 \tab 57 \tab 263 \tab 58 \tab 49 \tab 155 \tab 58 \tab 232 \tab 135 \tab 29 \tab 47 \tab 139 \tab 50 \tab 192 \tab 1196 \tab 228 \tab 242 \tab 162 \tab 1126 \tab 165 \tab 157 \tab 114 \tab 150 \tab 57 \tab 413 \tab 177 \tab 223 \tab 110 \tab 148 \tab 125 \tab 10 \tab 35 \tab 25 \tab 147 \tab 59 \tab 66 \tab 65 \tab 132 \tab 77 \tab 138 \tab 173 \tab 83 \tab 40 \tab 75 \tab 22 \tab 76 \tab 26 \tab 72 \tab 29 \tab 87 \tab 24 \tab 46 \tab 57 \tab 13 \tab 53 \tab 54 \tab 43 \tab 98 \tab 85 \tab 253 \tab 35 \tab 30 \tab 9 \tab 17 \tab 26 \tab 6 \tab 57 \tab 39 \tab 58 \tab 5 \tab 30 \tab 111 \tab 162 \tab 41 \tab 19 \tab 41 \tab 32 \tab 17 \tab 13 \tab 21 \tab 17 \tab 1 \tab 21 \tab 25 \tab 17 \tab 27 \tab 45 \tab 30 \tab 40 \tab 17 \tab 64 \tab 20 \tab 12 \tab 133 \tab 28 \tab 15 \tab 28 \tab 25 \tab 18 \tab 1 \tab 1 \tab 19 \tab 5 \tab 19 \tab 48 \tab 6 \tab 19 \tab 6 \tab 7 \tab 12 \tab 27 \tab 4 \tab 4 \tab 7 \tab 10 \tab 5 \tab 2 \tab 1 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{170} crop production
\item \code{180} animal production
\item \code{190} forestry, except logging
\item \code{270} logging
\item \code{280} fishing, hunting, and trapping
\item \code{290} support activities for agriculture and forestry
\item \code{370} oil and gas extraction
\item \code{380} coal mining
\item \code{390} metal ore mining
\item \code{470} nonmetallic mineral mining and quarrying
\item \code{480} not specified type of mining
\item \code{490} support activities for mining
\item \code{570} electric power generation, transmission, and distribution
\item \code{580} natural gas distribution
\item \code{590} electric and gas, and other combinations
\item \code{670} water, steam, air-conditioning, and irrigation systems
\item \code{680} sewage treatment facilities
\item \code{690} not specified utilities
\item \code{770} construction
\item \code{1070} animal food, grain, and oilseed milling
\item \code{1080} sugar and confectionery products
\item \code{1090} fruit and vegetable preserving and specialty food manufacturing
\item \code{1170} dairy product manufacturing
\item \code{1180} animal slaughtering and processing
\item \code{1190} retail bakeries
\item \code{1270} bakeries, except retail
\item \code{1280} seafood and other miscellaneous foods, n.e.c.
\item \code{1290} not specified food industries
\item \code{1370} beverage manufacturing
\item \code{1390} tobacco manufacturing
\item \code{1470} fiber, yarn, and thread mills
\item \code{1480} fabric mills, except knitting mills
\item \code{1490} textile and fabric finishing and coating mills
\item \code{1570} carpet and rug mills
\item \code{1590} textile product mills, except carpet and rug
\item \code{1670} knitting fabric mills, and apparel knitting mills
\item \code{1680} cut and sew apparel manufacturing
\item \code{1690} apparel accessories and other apparel manufacturing
\item \code{1770} footwear manufacturing
\item \code{1790} leather tanning and finishing and other allied products manufacturing
\item \code{1870} pulp, paper, and paperboard mills
\item \code{1880} paperboard containers and boxes
\item \code{1890} miscellaneous paper and pulp products
\item \code{1990} printing and related support activities
\item \code{2070} petroleum refining
\item \code{2090} miscellaneous petroleum and coal products
\item \code{2170} resin, synthetic rubber and fibers, and filaments manufacturing
\item \code{2180} agricultural chemical manufacturing
\item \code{2190} pharmaceutical and medicine manufacturing
\item \code{2270} paint, coating, and adhesive manufacturing
\item \code{2280} soap, cleaning compound, and cosmetics manufacturing
\item \code{2290} industrial and miscellaneous chemicals
\item \code{2370} plastics product manufacturing
\item \code{2380} tire manufacturing
\item \code{2390} rubber product, except tire, manufacturing
\item \code{2470} pottery, ceramics, and plumbing fixture manufacturing
\item \code{2480} structural clay product manufacturing
\item \code{2490} glass and glass product manufacturing
\item \code{2570} cement, concrete, lime, and gypsum product manufacturing
\item \code{2590} miscellaneous nonmetallic mineral product manufacturing
\item \code{2670} iron and steel mills and steel product manufacturing
\item \code{2680} aluminum production and processing
\item \code{2690} nonferrous metal (except aluminum) production and processing
\item \code{2770} foundries
\item \code{2780} metal forgings and stampings
\item \code{2790} cutlery and hand tool manufacturing
\item \code{2870} structural metals, and boiler, tank, and shipping container manufacturing
\item \code{2880} machine shops; turned product; screw, nut, and bolt manufacturing
\item \code{2890} coating, engraving, heat treating and allied activities
\item \code{2970} ordnance
\item \code{2980} miscellaneous fabricated metal products manufacturing
\item \code{2990} not specified metal industries
\item \code{3070} agricultural implement manufacturing
\item \code{3080} construction, and mining and oil and gas field machinery manufacturing
\item \code{3090} commercial and service industry machinery manufacturing
\item \code{3170} metalworking machinery manufacturing
\item \code{3180} engines, turbines, and power transmission equipment manufacturing
\item \code{3190} machinery manufacturing, n.e.c.
\item \code{3290} not specified machinery manufacturing
\item \code{3360} computer and peripheral equipment manufacturing
\item \code{3370} communications, and audio and video equipment manufacturing
\item \code{3380} navigational, measuring, electromedical, and control instruments manufacturing
\item \code{3390} electronic component and product manufacturing, n.e.c.
\item \code{3470} household appliance manufacturing
\item \code{3490} electrical lighting and electrical equipment manufacturing, and other electrical component manufacturing, n.e.c.
\item \code{3570} motor vehicles and motor vehicle equipment manufacturing
\item \code{3580} aircraft and parts manufacturing
\item \code{3590} aerospace product and parts manufacturing
\item \code{3670} railroad rolling stock manufacturing
\item \code{3680} ship and boat building
\item \code{3690} other transportation equipment manufacturing
\item \code{3770} sawmills and wood preservation
\item \code{3780} veneer, plywood, and engineered wood products
\item \code{3790} prefabricated wood buildings and mobile homes
\item \code{3870} miscellaneous wood products
\item \code{3890} furniture and related product manufacturing
\item \code{3960} medical equipment and supplies manufacturing
\item \code{3970} sporting and athletic goods, and doll, toy and game manufacturing
\item \code{3980} miscellaneous manufacturing, n.e.c.
\item \code{3990} not specified manufacturing industries
\item \code{4070} motor vehicles, parts and supplies, merchant wholesalers
\item \code{4080} furniture and home furnishing, merchant wholesalers
\item \code{4090} lumber and other construction materials, merchant wholesalers
\item \code{4170} professional and commercial equipment and supplies, merchant wholesalers
\item \code{4180} metals and minerals, except petroleum, merchant wholesalers
\item \code{4190} electrical and electronic goods, merchant wholesalers
\item \code{4260} hardware, plumbing and heating equipment, and supplies, merchant wholesalers
\item \code{4270} machinery, equipment, and supplies, merchant wholesalers
\item \code{4280} recyclable material, merchant wholesalers
\item \code{4290} miscellaneous durable goods, merchant wholesalers
\item \code{4370} paper and paper products, merchant wholesalers
\item \code{4380} drugs, sundries, and chemical and allied products, merchant wholesalers
\item \code{4390} apparel, fabrics, and notions, merchant wholesalers
\item \code{4470} groceries and related products, merchant wholesalers
\item \code{4480} farm product raw materials, merchant wholesalers
\item \code{4490} petroleum and petroleum products, merchant wholesalers
\item \code{4560} alcoholic beverages, merchant wholesalers
\item \code{4570} farm supplies, merchant wholesalers
\item \code{4580} miscellaneous nondurable goods, merchant wholesalers
\item \code{4585} wholesale electronic markets, agents and brokers
\item \code{4590} not specified wholesale trade
\item \code{4670} automobile dealers
\item \code{4680} other motor vehicle dealers
\item \code{4690} auto parts, accessories, and tire stores
\item \code{4770} furniture and home furnishings stores
\item \code{4780} household appliance stores
\item \code{4790} radio, tv, and computer stores
\item \code{4870} building material and supplies dealers
\item \code{4880} hardware stores
\item \code{4890} lawn and garden equipment and supplies stores
\item \code{4970} grocery stores
\item \code{4980} specialty food stores
\item \code{4990} beer, wine, and liquor stores
\item \code{5070} pharmacies and drug stores
\item \code{5080} health and personal care, except drug, stores
\item \code{5090} gasoline stations
\item \code{5170} clothing stores
\item \code{5180} shoe stores
\item \code{5190} jewelry, luggage, and leather goods stores
\item \code{5270} sporting goods, camera, and hobby and toy stores
\item \code{5280} sewing, needlework, and piece goods stores
\item \code{5290} music stores
\item \code{5370} book stores and news dealers
\item \code{5380} department stores and discount stores
\item \code{5390} miscellaneous general merchandise stores
\item \code{5470} retail florists
\item \code{5480} office supplies and stationery stores
\item \code{5490} used merchandise stores
\item \code{5570} gift, novelty, and souvenir shops
\item \code{5580} miscellaneous retail stores
\item \code{5590} electronic shopping
\item \code{5591} electronic auctions
\item \code{5592} mail order houses
\item \code{5670} vending machine operators
\item \code{5680} fuel dealers
\item \code{5690} other direct selling establishments
\item \code{5790} not specified retail trade
\item \code{6070} air transportation
\item \code{6080} rail transportation
\item \code{6090} water transportation
\item \code{6170} truck transportation
\item \code{6180} bus service and urban transit
\item \code{6190} taxi and limousine service
\item \code{6270} pipeline transportation
\item \code{6280} scenic and sightseeing transportation
\item \code{6290} services incidental to transportation
\item \code{6370} postal service
\item \code{6380} couriers and messengers
\item \code{6390} warehousing and storage
\item \code{6470} newspaper publishers
\item \code{6480} periodical, book, and directory publishers
\item \code{6490} software publishers
\item \code{6570} motion pictures and video industries
\item \code{6590} sound recording industries
\item \code{6670} radio and television broadcasting and cable subscription programming
\item \code{6672} internet publishing and broadcasting and web search portals
\item \code{6680} wired telecommunications carriers
\item \code{6690} other telecommunications services
\item \code{6695} data processing, hosting, and related services
\item \code{6770} libraries and archives
\item \code{6780} other information services
\item \code{6870} banking and related activities
\item \code{6880} savings institutions, including credit unions
\item \code{6890} non-depository credit and related activities
\item \code{6970} securities, commodities, funds, trusts, and other financial investments
\item \code{6990} insurance carriers and related activities
\item \code{7070} real estate
\item \code{7080} automotive equipment rental and leasing
\item \code{7170} video tape and disk rental
\item \code{7180} other consumer goods rental
\item \code{7190} commercial, industrial, and other intangible assets rental and leasing
\item \code{7270} legal services
\item \code{7280} accounting, tax preparation, bookkeeping, and payroll services
\item \code{7290} architectural, engineering, and related services
\item \code{7370} specialized design services
\item \code{7380} computer systems design and related services
\item \code{7390} management, scientific, and technical consulting services
\item \code{7460} scientific research and development services
\item \code{7470} advertising and related services
\item \code{7480} veterinary services
\item \code{7490} other professional, scientific, and technical services
\item \code{7570} management of companies and enterprises
\item \code{7580} employment services
\item \code{7590} business support services
\item \code{7670} travel arrangements and reservation services
\item \code{7680} investigation and security services
\item \code{7690} services to buildings and dwellings
\item \code{7770} landscaping services
\item \code{7780} other administrative and other support services
\item \code{7790} waste management and remediation services
\item \code{7860} elementary and secondary schools
\item \code{7870} colleges and universities, including junior colleges
\item \code{7880} business, technical, and trade schools and training
\item \code{7890} other schools and instruction, and educational support services
\item \code{7970} offices of physicians
\item \code{7980} offices of dentists
\item \code{7990} offices of chiropractors
\item \code{8070} offices of optometrists
\item \code{8080} offices of other health practitioners
\item \code{8090} outpatient care centers
\item \code{8170} home health care services
\item \code{8180} other health care services
\item \code{8190} hospitals
\item \code{8270} nursing care facilities
\item \code{8290} residential care facilities, without nursing
\item \code{8370} individual and family services
\item \code{8380} community food and housing, and emergency services
\item \code{8390} vocational rehabilitation services
\item \code{8470} child day care services
\item \code{8560} independent artists, performing arts, spectator sports, and related industries
\item \code{8570} museums, art galleries, historical sites, and similar institutions
\item \code{8580} bowling centers
\item \code{8590} other amusement, gambling, and recreation industries
\item \code{8660} traveler accommodation
\item \code{8670} recreational vehicle parks and camps, and rooming and boarding houses
\item \code{8680} restaurants and other food services
\item \code{8690} drinking places, alcoholic beverages
\item \code{8770} automotive repair and maintenance
\item \code{8780} car washes
\item \code{8790} electronic and precision equipment repair and maintenance
\item \code{8870} commercial and industrial machinery and equipment repair and maintenance
\item \code{8880} personal and household goods repair and maintenance
\item \code{8890} footwear and leather goods repair
\item \code{8970} barber shops
\item \code{8980} beauty salons
\item \code{8990} nail salons and other personal care services
\item \code{9070} drycleaning and laundry services
\item \code{9080} funeral homes, and cemeteries and crematories
\item \code{9090} other personal services
\item \code{9160} religious organizations
\item \code{9170} civic, social, advocacy organizations, and grantmaking and giving services
\item \code{9180} labor unions
\item \code{9190} business, professional, political, and similar organizations
\item \code{9290} private households
\item \code{9370} executive offices and legislative bodies
\item \code{9380} public finance activities
\item \code{9390} other general government and support
\item \code{9470} justice, public order, and safety activities
\item \code{9480} administration of human resource programs
\item \code{9490} administration of environmental quality and housing programs
\item \code{9570} administration of economic programs and space research
\item \code{9590} national security and international affairs
\item \code{9670} u.s. army
\item \code{9680} u.s. air force
\item \code{9690} u.s. navy
\item \code{9770} u.s. marine corps
\item \code{9780} u.s. coast guard
\item \code{9790} u.s. armed forces/n.s.
\item \code{9870} active duty military personnel, all other
\item \code{9890} uncodeable
\item \code{9997} uncodeable
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
