% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_07.R
\name{paind16}
\alias{paind16}
\title{paind16}
\source{
General Social Survey https://gss.norc.org
}
\description{
Father's industry code (1970)
}
\details{
Question 8c. What kind of place did he work for? d. What did they (make/do)?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab NA(i) \tab NA(n) \tab accounting, auditing, and bookkeeping services \tab agricultural chemicals \tab agricultural production \tab agricultural services, except horticultural \tab air transportation \tab aircraft and parts \tab alcoholic beverages \tab apparel and accessories \tab apparel and accessories stores, except shoe stores \tab automobile repair and related services \tab bakery products \tab banking \tab barber shops \tab beverage industries \tab blast furnaces, steel works, rolling and finishing mills \tab business management and consulting services \tab canning and preserving fruits, vegetables, and sea foods \tab coal mining \tab colleges and universities \tab construction and material handling machines \tab credit agencies \tab crude petroleum and natural gas extractions \tab cutlery, hand tools, and other hardware \tab dairy product stores \tab dairy products \tab department and mail order establishments \tab detective and protective services \tab direct selling establishments \tab drug stores \tab drugs and medicines \tab drugs, chemicals, and allied products \tab dry goods and apparel \tab dyeing and finishing textiles, except wool and knit goods \tab eating and drinking places \tab educational services, n.e.c. \tab electric light and power \tab electrical machinery, equipment, and supplies, n.e.c. \tab electrical repair shops \tab electronic computing equipment \tab elementary and secondary schools \tab engineering and architectural services \tab engines and turbines \tab fabricated structural metal products \tab farm products-raw materials \tab federal public administration \tab fisheries \tab floor coverings, except hard surface \tab food and related products \tab food stores, n.e.c. \tab footwear, except rubber \tab forestry \tab fuel and ice dealers \tab furniture and home furnishing stores \tab furniture fixtures \tab gas and steam supply systems \tab gasoline service stations \tab general building contractors \tab general contractors, except building \tab glass and glass products \tab grain-mill products \tab grocery stores \tab hardware and farm equipment stores \tab health services, n.e.c. \tab horticultural services \tab hospitals \tab hotels and motels \tab household appliances \tab household appliances, tv, and radio stores \tab industrial chemicals \tab insurance \tab jewelry stores \tab knitting mills \tab laundering, cleaning, and other garment services \tab leather products, except footwear \tab legal services \tab local public administration \tab lodging places, except hotels and motels \tab logging \tab lumber and building material retailing \tab lumber and construction materials \tab machinery equipment and supplies \tab machinery, except electrical, n.e.c. \tab meat products \tab metal mining \tab metals and minerals, n.e.c. \tab metalworking machinery \tab misc. general merchandise stores \tab miscellaneous chemicals \tab miscellaneous entertainment and recreation services \tab miscellaneous fabricated metal products \tab miscellaneous food preparation and kindred products \tab miscellaneous manufacturing industries \tab miscellaneous nonmetallic mineral and stone products \tab miscellaneous paper and pulp products \tab miscellaneous personal services \tab miscellaneous repair services \tab miscellaneous retail stores \tab miscellaneous textile mill products \tab miscellaneous wood products \tab motor vehicle dealers \tab motor vehicles and motor vehicle equipment \tab museums, art galleries, and zoos \tab newspaper publishing and printing \tab nonmetallic mining and quarrying, except fuel \tab not specified chemicals and allied products \tab not specified construction \tab not specified electrical machinery, equipment, and supplies \tab not specified food industries \tab not specified machinery \tab not specified metal industries \tab not specified retail trade \tab office and accounting machines \tab offices of chiropractors \tab offices of dentists \tab offices of physicians \tab optical and health services supplies \tab ordinance \tab other primary iron and steel industries \tab other primary nonferrous industries \tab paints, varnishes, and related products \tab paper and its products \tab paperboard containers and boxes \tab petroleum products \tab petroleum refining \tab photographic equipment and supplies \tab postal service \tab printing, publishing, and allied industries, except newspapers \tab private households \tab pulp, paper and paperboard mills \tab radio broadcasting and television \tab radio, t.v., and communication equipment \tab railroad locomotives and equipment \tab railroads and railway express \tab real estate, incl. real estate-insurance-law offices \tab religious organizations \tab retail bakeries \tab rubber products \tab sanitary services \tab sawmills, planing mills, and  mill work \tab scientific and controlling instruments \tab scrap and waste materials \tab security, commodity brokerage, and investment companies \tab services to dwellings and other buildings \tab ship and boat building and repairing \tab shoe repair shops \tab soaps and cosmetics \tab special trade contractors \tab state public administration \tab street railways and bus lines \tab structural clay products \tab tanned, curried, and finished leather \tab taxicab service \tab telegraph and miscellaneous communication services \tab telephone (wire and radio) \tab theaters and motion pictures \tab tire, battery, and accessory dealers \tab tobacco manufactures \tab trucking service \tab warehousing and storage \tab water supply \tab water transportation \tab welfare services \tab wholesalers, n.e.c. \tab yarn, thread, and fabric mills \tab advertising \tab business services, n.e.c. \tab cement concrete, gypsum, and  plaster products \tab commercial research, development, and testing labs \tab electric-gas utilities \tab electrical goods \tab farm machinery and equipment \tab hardware, plumbing, and heating supplies \tab miscellaneous professional and related services \tab miscellaneous vehicle dealers \tab motor vehicles and equipment \tab nonprofit membership organizations \tab offices of health practitioners,n.e.c. \tab plastics, synthetics and resins, except fibers \tab primary aluminum industries \tab retail florists \tab screw machine products \tab services incidental to transportation \tab shoe stores \tab synthetic fibers \tab automobile services, except repair \tab bowling alleys, billiard and pool parlors \tab confectionery and related products \tab limited price variety stores \tab liquor stores \tab miscellaneous fabricated textile products \tab miscellaneous petroleum and coal products \tab not specified manufacturing industries \tab not specified professional equipment \tab other and not specified utilities \tab pipe lines, except natural gas \tab pottery and related products \tab watches, clocks, and clockwork-operated devices \tab convalescent institutions \tab farm and garden supply stores \tab metal stamping \tab miscellaneous plastic products \tab vending machine operators \tab beauty shops \tab cycles and miscellaneous transportation equipment \tab mobile dwellings and campers \tab computer programming services \tab not specified wholesale trade \tab residential welfare facilities \tab dressmaking shops \tab employment and temporary help agencies \tab not specified educational services \tab not specified electrical and hardware products \tab Total \cr
   1972 \tab 198 \tab 87 \tab 1 \tab 1 \tab 360 \tab 5 \tab 3 \tab 8 \tab 5 \tab 12 \tab 8 \tab 13 \tab 7 \tab 8 \tab 12 \tab 7 \tab 20 \tab 2 \tab 1 \tab 33 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 5 \tab 3 \tab 2 \tab 4 \tab 1 \tab 14 \tab 1 \tab 8 \tab 3 \tab 1 \tab 3 \tab 17 \tab 1 \tab 1 \tab 3 \tab 5 \tab 26 \tab 1 \tab 1 \tab 15 \tab 5 \tab 3 \tab 2 \tab 2 \tab 2 \tab 5 \tab 2 \tab 8 \tab 57 \tab 24 \tab 1 \tab 5 \tab 14 \tab 1 \tab 1 \tab 2 \tab 7 \tab 6 \tab 3 \tab 2 \tab 5 \tab 14 \tab 4 \tab 1 \tab 1 \tab 2 \tab 2 \tab 27 \tab 1 \tab 8 \tab 10 \tab 2 \tab 3 \tab 7 \tab 5 \tab 8 \tab 1 \tab 2 \tab 8 \tab 3 \tab 3 \tab 8 \tab 1 \tab 4 \tab 1 \tab 3 \tab 3 \tab 5 \tab 2 \tab 3 \tab 1 \tab 7 \tab 14 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 6 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 6 \tab 1 \tab 7 \tab 3 \tab 8 \tab 2 \tab 1 \tab 1 \tab 3 \tab 4 \tab 1 \tab 15 \tab 10 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 46 \tab 11 \tab 11 \tab 4 \tab 14 \tab 5 \tab 18 \tab 1 \tab 4 \tab 3 \tab 1 \tab 8 \tab 1 \tab 1 \tab 50 \tab 5 \tab 6 \tab 2 \tab 5 \tab 4 \tab 1 \tab 8 \tab 1 \tab 1 \tab 1 \tab 20 \tab 4 \tab 4 \tab 9 \tab 1 \tab 4 \tab 13 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 149 \tab 25 \tab 2 \tab 1 \tab 336 \tab 2 \tab 3 \tab 5 \tab 2 \tab 3 \tab 6 \tab 5 \tab 3 \tab 9 \tab 5 \tab 4 \tab 23 \tab - \tab 3 \tab 43 \tab 5 \tab 3 \tab 1 \tab 6 \tab 5 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 7 \tab 3 \tab 1 \tab 3 \tab 1 \tab 15 \tab 1 \tab 8 \tab 3 \tab - \tab 1 \tab 8 \tab 3 \tab - \tab 5 \tab 1 \tab 29 \tab 2 \tab - \tab 13 \tab 1 \tab 2 \tab - \tab 3 \tab 6 \tab 6 \tab 1 \tab 6 \tab 51 \tab 28 \tab 5 \tab 3 \tab 23 \tab 4 \tab 1 \tab 3 \tab 7 \tab 2 \tab 8 \tab 8 \tab 2 \tab 14 \tab 3 \tab 2 \tab 3 \tab - \tab 2 \tab 19 \tab 2 \tab 3 \tab 10 \tab 4 \tab 4 \tab 16 \tab 2 \tab 10 \tab 2 \tab 2 \tab 8 \tab 4 \tab 2 \tab 8 \tab 6 \tab 5 \tab 1 \tab 2 \tab 4 \tab 6 \tab 4 \tab 1 \tab 2 \tab 10 \tab 19 \tab - \tab 6 \tab 2 \tab 1 \tab 6 \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 5 \tab 1 \tab 2 \tab 4 \tab 4 \tab 1 \tab - \tab 3 \tab 1 \tab 3 \tab 1 \tab 10 \tab 11 \tab 2 \tab 8 \tab 1 \tab 3 \tab 1 \tab 68 \tab 9 \tab 14 \tab 4 \tab 4 \tab 3 \tab 22 \tab - \tab 6 \tab - \tab 1 \tab 9 \tab 2 \tab - \tab 52 \tab 3 \tab 4 \tab - \tab - \tab 8 \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab 23 \tab 3 \tab 5 \tab 7 \tab 1 \tab 3 \tab 7 \tab 4 \tab 2 \tab 6 \tab 1 \tab 1 \tab 2 \tab 5 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 182 \tab 20 \tab 1 \tab 1 \tab 309 \tab - \tab 1 \tab 10 \tab - \tab 5 \tab 7 \tab 9 \tab 4 \tab 3 \tab 9 \tab 6 \tab 2 \tab - \tab - \tab 38 \tab 4 \tab 3 \tab - \tab 5 \tab 1 \tab 1 \tab 6 \tab 4 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 14 \tab - \tab 3 \tab 17 \tab 1 \tab - \tab 11 \tab - \tab 2 \tab 2 \tab 2 \tab 28 \tab - \tab 3 \tab 8 \tab 7 \tab 2 \tab - \tab 1 \tab 6 \tab 7 \tab 2 \tab 5 \tab 76 \tab 14 \tab 6 \tab - \tab 15 \tab 7 \tab 1 \tab 3 \tab 8 \tab 4 \tab 2 \tab 3 \tab 12 \tab 9 \tab 2 \tab - \tab 4 \tab - \tab 3 \tab 16 \tab - \tab 11 \tab 4 \tab 2 \tab 2 \tab 5 \tab 3 \tab 12 \tab - \tab 1 \tab 7 \tab - \tab 6 \tab 4 \tab 5 \tab 13 \tab 1 \tab 4 \tab 10 \tab 5 \tab 5 \tab 4 \tab 5 \tab 7 \tab 31 \tab 1 \tab 9 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab 4 \tab 1 \tab 2 \tab 16 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 14 \tab 5 \tab 3 \tab 2 \tab - \tab - \tab - \tab 58 \tab 9 \tab 8 \tab 4 \tab 10 \tab 4 \tab 14 \tab 1 \tab - \tab 6 \tab - \tab 7 \tab - \tab 3 \tab 48 \tab 15 \tab 7 \tab - \tab 1 \tab 1 \tab 2 \tab 6 \tab 4 \tab 2 \tab 6 \tab 11 \tab 1 \tab 7 \tab 2 \tab - \tab 2 \tab 10 \tab 1 \tab 1 \tab 5 \tab - \tab 2 \tab 3 \tab 7 \tab 4 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 151 \tab 25 \tab 2 \tab - \tab 290 \tab 2 \tab 3 \tab 13 \tab - \tab 14 \tab 4 \tab 13 \tab 2 \tab 6 \tab 9 \tab 7 \tab 20 \tab - \tab 1 \tab 36 \tab 8 \tab 2 \tab 3 \tab 7 \tab 5 \tab - \tab 12 \tab 10 \tab 1 \tab 2 \tab 6 \tab 2 \tab 1 \tab - \tab 1 \tab 13 \tab - \tab 9 \tab 3 \tab - \tab 2 \tab 10 \tab 3 \tab - \tab 2 \tab 2 \tab 39 \tab 4 \tab - \tab 9 \tab 5 \tab 2 \tab 3 \tab 4 \tab 9 \tab 10 \tab 2 \tab 10 \tab 48 \tab 26 \tab 5 \tab 4 \tab 23 \tab 4 \tab 1 \tab 1 \tab 4 \tab 7 \tab 5 \tab 5 \tab 1 \tab 12 \tab 2 \tab 4 \tab 6 \tab - \tab 1 \tab 23 \tab - \tab 5 \tab 5 \tab 1 \tab 2 \tab 14 \tab 6 \tab 14 \tab 1 \tab 3 \tab 2 \tab 6 \tab 4 \tab 1 \tab 3 \tab 4 \tab 3 \tab 5 \tab - \tab 6 \tab 1 \tab - \tab 1 \tab 15 \tab 27 \tab - \tab 5 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab 8 \tab 2 \tab 2 \tab 7 \tab 7 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 16 \tab 6 \tab 3 \tab 1 \tab 2 \tab 3 \tab 2 \tab 46 \tab 13 \tab 9 \tab - \tab 4 \tab 7 \tab 20 \tab 2 \tab - \tab - \tab 1 \tab 11 \tab 3 \tab - \tab 50 \tab 7 \tab 4 \tab 4 \tab 3 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 27 \tab 3 \tab 3 \tab 7 \tab - \tab 4 \tab 8 \tab 3 \tab 2 \tab 5 \tab - \tab 3 \tab 1 \tab 2 \tab 2 \tab 4 \tab - \tab 2 \tab 5 \tab 1 \tab - \tab 4 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 187 \tab 22 \tab 1 \tab 4 \tab 302 \tab 3 \tab - \tab 19 \tab 1 \tab 10 \tab 6 \tab 13 \tab 6 \tab 6 \tab 2 \tab 4 \tab 35 \tab - \tab 5 \tab 47 \tab 5 \tab 3 \tab - \tab 8 \tab 6 \tab - \tab 10 \tab 5 \tab 1 \tab 5 \tab - \tab 2 \tab - \tab - \tab 1 \tab 15 \tab 1 \tab 7 \tab 4 \tab 2 \tab 1 \tab 15 \tab 1 \tab 1 \tab 8 \tab 1 \tab 29 \tab 1 \tab 1 \tab 7 \tab 4 \tab 3 \tab - \tab 4 \tab 5 \tab 4 \tab 5 \tab 2 \tab 39 \tab 16 \tab 6 \tab 1 \tab 16 \tab 5 \tab - \tab 3 \tab 8 \tab 3 \tab 4 \tab 2 \tab 3 \tab 16 \tab - \tab 1 \tab 7 \tab - \tab 4 \tab 23 \tab - \tab 5 \tab 6 \tab 3 \tab 1 \tab 12 \tab 5 \tab 18 \tab 2 \tab 7 \tab 2 \tab - \tab 3 \tab 2 \tab 4 \tab 7 \tab 2 \tab 1 \tab 2 \tab 3 \tab 5 \tab - \tab 1 \tab 11 \tab 33 \tab - \tab 6 \tab 3 \tab 2 \tab 3 \tab 3 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 2 \tab 9 \tab 1 \tab 4 \tab 10 \tab 4 \tab - \tab 2 \tab 4 \tab 2 \tab - \tab 2 \tab 7 \tab 8 \tab 5 \tab 3 \tab 3 \tab 5 \tab 1 \tab 55 \tab 15 \tab 9 \tab 2 \tab 4 \tab 6 \tab 19 \tab 1 \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 38 \tab 8 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab 9 \tab 2 \tab 1 \tab 1 \tab 19 \tab 2 \tab 2 \tab 5 \tab - \tab 1 \tab 6 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 6 \tab 3 \tab - \tab - \tab 3 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 193 \tab 17 \tab 2 \tab - \tab 294 \tab 4 \tab 4 \tab 10 \tab 1 \tab 2 \tab 4 \tab 10 \tab 8 \tab 8 \tab 4 \tab 9 \tab 29 \tab 1 \tab - \tab 54 \tab 7 \tab 6 \tab 2 \tab 13 \tab 3 \tab - \tab 9 \tab 3 \tab - \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab 15 \tab 1 \tab 11 \tab 8 \tab 1 \tab 3 \tab 12 \tab 2 \tab 1 \tab 5 \tab 3 \tab 42 \tab 2 \tab 2 \tab 5 \tab 3 \tab 2 \tab - \tab 3 \tab 5 \tab 5 \tab 1 \tab 6 \tab 34 \tab 25 \tab 2 \tab 3 \tab 15 \tab 8 \tab - \tab 6 \tab 1 \tab 8 \tab 3 \tab 1 \tab 1 \tab 10 \tab 2 \tab - \tab 8 \tab 3 \tab 9 \tab 22 \tab - \tab 7 \tab 2 \tab 3 \tab 5 \tab 15 \tab - \tab 19 \tab - \tab 7 \tab 4 \tab - \tab 1 \tab 6 \tab 5 \tab 7 \tab 4 \tab 1 \tab 1 \tab 8 \tab 3 \tab - \tab 2 \tab 5 \tab 26 \tab 1 \tab 7 \tab 1 \tab 2 \tab 11 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 4 \tab - \tab 1 \tab 11 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab 16 \tab 3 \tab 5 \tab 6 \tab 1 \tab 3 \tab 3 \tab 45 \tab 5 \tab 9 \tab 4 \tab 7 \tab 5 \tab 23 \tab - \tab 4 \tab 4 \tab 2 \tab 11 \tab - \tab 1 \tab 52 \tab 2 \tab 5 \tab 5 \tab 2 \tab 1 \tab 1 \tab 5 \tab 4 \tab 4 \tab 1 \tab 21 \tab 2 \tab 7 \tab 8 \tab - \tab 2 \tab 10 \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab - \tab 7 \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 205 \tab 9 \tab 1 \tab - \tab 262 \tab 4 \tab 3 \tab 18 \tab 2 \tab 6 \tab 3 \tab 12 \tab 6 \tab 6 \tab 4 \tab 4 \tab 23 \tab 1 \tab 5 \tab 57 \tab 2 \tab 3 \tab 1 \tab 8 \tab 4 \tab - \tab 10 \tab 7 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 16 \tab - \tab 8 \tab 7 \tab 1 \tab 1 \tab 16 \tab 5 \tab - \tab 6 \tab 5 \tab 47 \tab 1 \tab - \tab 8 \tab 3 \tab 3 \tab 3 \tab 2 \tab 1 \tab 13 \tab 4 \tab 4 \tab 52 \tab 22 \tab 7 \tab 5 \tab 15 \tab 4 \tab 2 \tab 1 \tab 7 \tab 2 \tab - \tab 4 \tab 3 \tab 13 \tab 2 \tab 1 \tab 6 \tab 2 \tab 8 \tab 34 \tab - \tab 11 \tab 3 \tab 3 \tab 1 \tab 7 \tab 4 \tab 20 \tab 1 \tab 6 \tab 4 \tab 3 \tab - \tab 5 \tab 3 \tab 7 \tab 1 \tab 2 \tab 1 \tab 9 \tab 3 \tab 2 \tab 2 \tab 11 \tab 28 \tab - \tab 6 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 6 \tab 1 \tab 4 \tab 9 \tab 5 \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab 2 \tab 15 \tab 7 \tab 2 \tab 4 \tab 3 \tab 2 \tab 2 \tab 43 \tab 5 \tab 7 \tab 8 \tab 4 \tab 2 \tab 14 \tab 1 \tab 5 \tab 3 \tab - \tab 8 \tab 2 \tab 3 \tab 57 \tab 10 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 3 \tab 17 \tab 2 \tab 2 \tab 5 \tab - \tab 2 \tab 13 \tab 3 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 4 \tab - \tab - \tab 6 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 5 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 230 \tab 9 \tab - \tab 1 \tab 244 \tab 3 \tab 2 \tab 10 \tab 3 \tab 9 \tab 4 \tab 11 \tab 5 \tab 5 \tab 9 \tab 2 \tab 21 \tab 1 \tab 3 \tab 55 \tab 10 \tab - \tab - \tab 12 \tab 6 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 4 \tab - \tab 12 \tab - \tab 7 \tab 1 \tab - \tab 2 \tab 18 \tab 2 \tab 2 \tab 5 \tab 2 \tab 47 \tab 2 \tab 1 \tab 9 \tab 2 \tab 1 \tab - \tab 1 \tab 5 \tab 6 \tab 3 \tab 8 \tab 34 \tab 11 \tab 6 \tab 4 \tab 20 \tab 1 \tab - \tab 1 \tab 7 \tab 5 \tab 1 \tab 3 \tab 5 \tab 12 \tab 3 \tab 2 \tab 4 \tab 2 \tab 7 \tab 27 \tab - \tab 8 \tab 5 \tab 1 \tab 1 \tab 12 \tab 4 \tab 14 \tab 1 \tab - \tab 2 \tab 3 \tab 4 \tab 11 \tab 2 \tab 11 \tab 1 \tab 3 \tab 4 \tab 7 \tab 3 \tab 5 \tab 4 \tab 10 \tab 21 \tab - \tab 9 \tab 5 \tab 4 \tab 13 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab - \tab 7 \tab 3 \tab - \tab 6 \tab 3 \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab 2 \tab 8 \tab 9 \tab 4 \tab 7 \tab - \tab 2 \tab 4 \tab 43 \tab 9 \tab 6 \tab 1 \tab 5 \tab 2 \tab 17 \tab 1 \tab 1 \tab 2 \tab 5 \tab 5 \tab 1 \tab 2 \tab 49 \tab 10 \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab 10 \tab 1 \tab 3 \tab - \tab 17 \tab 3 \tab - \tab 3 \tab - \tab 3 \tab 8 \tab 1 \tab 2 \tab 1 \tab 1 \tab 8 \tab - \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab 3 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 296 \tab 65 \tab - \tab 1 \tab 304 \tab 6 \tab 2 \tab 10 \tab 5 \tab 11 \tab 5 \tab 15 \tab 2 \tab 8 \tab 8 \tab 6 \tab 29 \tab - \tab 5 \tab 47 \tab 4 \tab 8 \tab 1 \tab 6 \tab 2 \tab - \tab 6 \tab 4 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 5 \tab - \tab 22 \tab 1 \tab 8 \tab 4 \tab 2 \tab 1 \tab 24 \tab 6 \tab 2 \tab 10 \tab 3 \tab 43 \tab 3 \tab - \tab 15 \tab 5 \tab 2 \tab 6 \tab 1 \tab 4 \tab 4 \tab 6 \tab 10 \tab 50 \tab 38 \tab 4 \tab 3 \tab 23 \tab 6 \tab - \tab 1 \tab 10 \tab 4 \tab 1 \tab 2 \tab - \tab 11 \tab 2 \tab 1 \tab 4 \tab 2 \tab 3 \tab 26 \tab - \tab 11 \tab 2 \tab 1 \tab 4 \tab 7 \tab 6 \tab 13 \tab - \tab 9 \tab 7 \tab 1 \tab 4 \tab 5 \tab 1 \tab 5 \tab - \tab 3 \tab 3 \tab 3 \tab 3 \tab 2 \tab 2 \tab 13 \tab 24 \tab 1 \tab 9 \tab 2 \tab 5 \tab 10 \tab 2 \tab 2 \tab 1 \tab 8 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 3 \tab 3 \tab 7 \tab 4 \tab 9 \tab 1 \tab 1 \tab 7 \tab 6 \tab 5 \tab 18 \tab 7 \tab 6 \tab 3 \tab - \tab 2 \tab 2 \tab 47 \tab 14 \tab 14 \tab 4 \tab 7 \tab 4 \tab 21 \tab - \tab 2 \tab 1 \tab 1 \tab 14 \tab 3 \tab 2 \tab 66 \tab 7 \tab 14 \tab 4 \tab 1 \tab 2 \tab - \tab 6 \tab 3 \tab 2 \tab 1 \tab 35 \tab 2 \tab 2 \tab 9 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 3 \tab - \tab 4 \tab 1 \tab 5 \tab 3 \tab 2 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 207 \tab 11 \tab 3 \tab - \tab 215 \tab 2 \tab 3 \tab 10 \tab - \tab 14 \tab 8 \tab 15 \tab - \tab 10 \tab 10 \tab 1 \tab 43 \tab 1 \tab 1 \tab 21 \tab 12 \tab 4 \tab 2 \tab 9 \tab 3 \tab - \tab 11 \tab 11 \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 9 \tab - \tab 12 \tab 4 \tab 3 \tab 2 \tab 23 \tab 6 \tab 3 \tab 24 \tab 1 \tab 37 \tab 4 \tab - \tab 15 \tab 6 \tab 6 \tab - \tab 5 \tab 7 \tab 15 \tab - \tab 7 \tab 78 \tab 7 \tab 1 \tab 4 \tab 22 \tab 4 \tab 1 \tab 3 \tab 5 \tab 5 \tab 2 \tab 3 \tab 4 \tab 11 \tab 1 \tab 5 \tab 6 \tab 1 \tab 9 \tab 31 \tab - \tab 5 \tab 6 \tab - \tab 3 \tab 1 \tab 10 \tab 21 \tab 1 \tab 2 \tab 3 \tab 2 \tab 3 \tab 6 \tab 3 \tab 2 \tab 2 \tab 2 \tab - \tab 5 \tab 4 \tab - \tab 2 \tab 16 \tab 27 \tab - \tab 4 \tab 4 \tab 3 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 6 \tab 8 \tab - \tab 9 \tab 8 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 18 \tab 9 \tab 5 \tab 5 \tab 1 \tab 9 \tab - \tab 35 \tab 12 \tab 13 \tab 2 \tab 5 \tab 2 \tab 13 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 83 \tab 4 \tab 3 \tab 2 \tab - \tab 2 \tab - \tab 9 \tab 3 \tab 1 \tab 3 \tab 26 \tab - \tab 5 \tab 10 \tab - \tab 4 \tab 11 \tab 2 \tab 4 \tab 5 \tab - \tab 3 \tab 1 \tab 10 \tab 5 \tab 3 \tab 1 \tab 3 \tab 1 \tab 3 \tab 6 \tab 7 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 5 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 202 \tab 24 \tab 2 \tab - \tab 228 \tab 12 \tab 7 \tab 8 \tab - \tab 6 \tab 4 \tab 12 \tab 3 \tab 4 \tab 8 \tab 1 \tab 14 \tab 1 \tab 2 \tab 29 \tab 7 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 20 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 21 \tab - \tab 11 \tab 8 \tab - \tab 3 \tab 19 \tab 4 \tab 3 \tab 5 \tab 1 \tab 36 \tab 5 \tab - \tab 8 \tab 2 \tab 3 \tab 1 \tab 2 \tab 6 \tab 5 \tab 2 \tab 6 \tab 78 \tab 14 \tab 6 \tab 4 \tab 17 \tab 1 \tab 7 \tab 6 \tab 6 \tab 2 \tab 6 \tab 3 \tab 2 \tab 20 \tab 1 \tab 4 \tab 5 \tab 1 \tab 4 \tab 24 \tab - \tab 5 \tab 1 \tab 3 \tab - \tab 1 \tab 9 \tab 1 \tab - \tab 2 \tab - \tab 5 \tab 3 \tab 18 \tab 4 \tab 32 \tab 5 \tab 5 \tab 7 \tab 6 \tab 8 \tab 2 \tab 3 \tab 8 \tab 25 \tab - \tab 6 \tab 2 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 24 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 7 \tab - \tab 12 \tab 9 \tab 13 \tab 6 \tab 1 \tab 3 \tab 3 \tab 31 \tab 12 \tab 7 \tab - \tab 4 \tab 2 \tab 9 \tab 2 \tab 2 \tab 4 \tab 3 \tab 10 \tab - \tab - \tab 27 \tab 9 \tab 6 \tab 2 \tab 1 \tab 2 \tab 9 \tab 4 \tab 2 \tab 3 \tab 1 \tab 22 \tab 1 \tab 3 \tab 9 \tab - \tab 3 \tab 5 \tab 2 \tab 4 \tab 3 \tab - \tab 1 \tab 1 \tab 8 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 194 \tab 7 \tab 2 \tab - \tab 248 \tab 4 \tab 1 \tab 8 \tab 2 \tab 8 \tab 4 \tab 18 \tab 2 \tab 7 \tab 4 \tab 8 \tab 40 \tab 1 \tab - \tab 18 \tab 10 \tab 3 \tab 2 \tab 6 \tab 1 \tab - \tab 3 \tab 5 \tab - \tab 1 \tab 4 \tab 2 \tab 1 \tab 3 \tab - \tab 15 \tab 1 \tab 9 \tab 4 \tab 2 \tab 4 \tab 16 \tab 5 \tab 4 \tab 3 \tab 2 \tab 48 \tab 2 \tab - \tab 8 \tab 7 \tab 4 \tab - \tab 6 \tab 3 \tab 5 \tab 4 \tab 4 \tab 50 \tab 25 \tab 7 \tab 4 \tab 24 \tab 1 \tab - \tab 4 \tab 8 \tab 1 \tab 1 \tab 3 \tab 1 \tab 20 \tab - \tab 1 \tab 4 \tab - \tab 8 \tab 28 \tab - \tab 3 \tab 2 \tab 2 \tab 3 \tab 12 \tab 13 \tab 6 \tab 1 \tab 3 \tab 5 \tab 1 \tab 3 \tab 7 \tab 3 \tab 6 \tab 2 \tab 1 \tab 4 \tab 3 \tab 3 \tab 1 \tab 4 \tab 14 \tab 27 \tab - \tab 3 \tab 6 \tab 3 \tab 5 \tab 4 \tab 2 \tab 6 \tab 3 \tab - \tab - \tab - \tab 3 \tab 12 \tab 1 \tab 3 \tab 5 \tab 3 \tab 1 \tab - \tab 3 \tab 2 \tab 5 \tab 1 \tab 18 \tab 7 \tab 2 \tab 8 \tab 2 \tab 5 \tab 1 \tab 40 \tab 13 \tab 14 \tab 6 \tab 6 \tab 4 \tab 15 \tab 4 \tab 2 \tab 3 \tab 4 \tab 8 \tab - \tab 1 \tab 62 \tab 10 \tab 3 \tab 5 \tab 1 \tab 4 \tab - \tab 8 \tab - \tab 1 \tab 2 \tab 21 \tab 1 \tab 7 \tab 5 \tab 3 \tab 2 \tab 11 \tab 1 \tab 7 \tab 10 \tab 2 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 6 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 222 \tab 25 \tab - \tab - \tab 243 \tab 4 \tab 3 \tab 7 \tab 2 \tab 9 \tab - \tab 12 \tab 1 \tab 5 \tab 4 \tab 8 \tab 31 \tab - \tab 2 \tab 13 \tab 9 \tab 2 \tab 1 \tab 6 \tab 1 \tab - \tab 2 \tab 7 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 11 \tab - \tab 9 \tab 1 \tab 1 \tab 1 \tab 12 \tab 3 \tab - \tab 5 \tab 1 \tab 37 \tab 1 \tab 1 \tab 7 \tab 2 \tab 2 \tab - \tab 5 \tab 8 \tab 8 \tab 4 \tab 5 \tab 38 \tab 12 \tab 3 \tab 1 \tab 15 \tab 3 \tab 2 \tab 1 \tab 11 \tab 4 \tab 4 \tab 4 \tab 2 \tab 10 \tab 1 \tab 2 \tab 7 \tab - \tab 6 \tab 32 \tab - \tab 4 \tab 3 \tab 4 \tab 4 \tab 14 \tab 7 \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab 8 \tab 5 \tab 4 \tab 7 \tab 3 \tab 4 \tab 1 \tab 11 \tab 11 \tab 7 \tab 2 \tab 13 \tab 33 \tab 1 \tab 7 \tab 5 \tab 5 \tab 6 \tab 6 \tab - \tab 4 \tab 2 \tab 1 \tab 2 \tab - \tab 3 \tab 3 \tab 1 \tab 6 \tab 6 \tab 4 \tab - \tab - \tab 4 \tab 1 \tab 4 \tab 1 \tab 11 \tab 7 \tab 4 \tab 6 \tab 5 \tab 3 \tab 2 \tab 33 \tab 16 \tab 10 \tab 2 \tab 3 \tab - \tab 10 \tab 2 \tab 1 \tab 2 \tab 3 \tab 4 \tab 3 \tab 1 \tab 65 \tab 7 \tab 7 \tab 2 \tab - \tab 3 \tab - \tab 6 \tab - \tab 2 \tab 1 \tab 20 \tab 4 \tab 1 \tab 5 \tab - \tab 5 \tab 7 \tab 2 \tab 5 \tab 6 \tab - \tab 1 \tab 3 \tab 5 \tab 2 \tab 1 \tab - \tab 1 \tab 5 \tab - \tab 2 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab 2 \tab 1 \tab 1 \tab 11 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1470 \cr
   1987 \tab 291 \tab 26 \tab 3 \tab 3 \tab 306 \tab 16 \tab 7 \tab 11 \tab 3 \tab 7 \tab 4 \tab 12 \tab 6 \tab 6 \tab 2 \tab 2 \tab 38 \tab 2 \tab 2 \tab 31 \tab 14 \tab 2 \tab 4 \tab 6 \tab 2 \tab - \tab 8 \tab 7 \tab - \tab 2 \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 12 \tab - \tab 7 \tab 10 \tab 2 \tab 2 \tab 24 \tab 6 \tab 3 \tab 4 \tab 1 \tab 43 \tab 2 \tab 1 \tab 9 \tab 2 \tab 11 \tab 2 \tab 1 \tab 9 \tab 5 \tab 5 \tab 3 \tab 67 \tab 16 \tab 7 \tab 3 \tab 20 \tab 3 \tab 2 \tab 4 \tab 12 \tab 3 \tab 7 \tab 2 \tab 1 \tab 17 \tab 1 \tab 1 \tab 6 \tab 1 \tab 8 \tab 35 \tab 1 \tab 5 \tab 3 \tab 1 \tab 7 \tab 12 \tab 7 \tab 2 \tab - \tab 2 \tab 2 \tab 4 \tab 5 \tab 9 \tab 4 \tab 14 \tab 2 \tab 5 \tab - \tab 14 \tab 5 \tab 4 \tab 6 \tab 11 \tab 37 \tab - \tab 2 \tab 4 \tab 8 \tab 15 \tab 1 \tab 1 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab 3 \tab 10 \tab - \tab 1 \tab 6 \tab 6 \tab 2 \tab - \tab 5 \tab 3 \tab - \tab 1 \tab 12 \tab 12 \tab 2 \tab 6 \tab 2 \tab - \tab 1 \tab 50 \tab 7 \tab 13 \tab 1 \tab 13 \tab - \tab 9 \tab 1 \tab 2 \tab 2 \tab 2 \tab 5 \tab - \tab - \tab 66 \tab 11 \tab 10 \tab 2 \tab 2 \tab 6 \tab 2 \tab 4 \tab 3 \tab 5 \tab 1 \tab 30 \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab 9 \tab - \tab 6 \tab 4 \tab 1 \tab 6 \tab 4 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 4 \tab 8 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 207 \tab 19 \tab 1 \tab 1 \tab 225 \tab 3 \tab 7 \tab 11 \tab - \tab 3 \tab 2 \tab 14 \tab 1 \tab 2 \tab 4 \tab 2 \tab 4 \tab 2 \tab 1 \tab 13 \tab 15 \tab 1 \tab - \tab 4 \tab - \tab - \tab 4 \tab 3 \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab 4 \tab 13 \tab 2 \tab 6 \tab 6 \tab 4 \tab 4 \tab 28 \tab 7 \tab 3 \tab 10 \tab 3 \tab 55 \tab 2 \tab - \tab 9 \tab 4 \tab 4 \tab 2 \tab 3 \tab 2 \tab 11 \tab 1 \tab 8 \tab 54 \tab 5 \tab 1 \tab 2 \tab 21 \tab 8 \tab - \tab 4 \tab 9 \tab 4 \tab 2 \tab 3 \tab 3 \tab 10 \tab 3 \tab - \tab 4 \tab 1 \tab 4 \tab 52 \tab 1 \tab 5 \tab 7 \tab 3 \tab 6 \tab 12 \tab 7 \tab 4 \tab 1 \tab 3 \tab 1 \tab 5 \tab 4 \tab 5 \tab - \tab 4 \tab - \tab - \tab 3 \tab 2 \tab 4 \tab 6 \tab - \tab 10 \tab 26 \tab 1 \tab 7 \tab - \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab 5 \tab 3 \tab 4 \tab 20 \tab 4 \tab - \tab - \tab 2 \tab 5 \tab 8 \tab 4 \tab 9 \tab 5 \tab 1 \tab 6 \tab 5 \tab 4 \tab - \tab 31 \tab 6 \tab 12 \tab 7 \tab 4 \tab 2 \tab 12 \tab 1 \tab - \tab 1 \tab 6 \tab 4 \tab 2 \tab - \tab 61 \tab 14 \tab 9 \tab 1 \tab - \tab 6 \tab 2 \tab 9 \tab 2 \tab 4 \tab - \tab 24 \tab 7 \tab 1 \tab 3 \tab 2 \tab 2 \tab 7 \tab 3 \tab - \tab 4 \tab 1 \tab 7 \tab - \tab 5 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 11 \tab - \tab 2 \tab - \tab 5 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1481 \cr
   1989 \tab 225 \tab 23 \tab 4 \tab - \tab 205 \tab 30 \tab 5 \tab 5 \tab - \tab 3 \tab 2 \tab 18 \tab 2 \tab 6 \tab 5 \tab 4 \tab 23 \tab 4 \tab 3 \tab 25 \tab 15 \tab 6 \tab 2 \tab 7 \tab - \tab - \tab 1 \tab 5 \tab 1 \tab - \tab 4 \tab 1 \tab 3 \tab - \tab 2 \tab 19 \tab - \tab 9 \tab 7 \tab 3 \tab 3 \tab 26 \tab 9 \tab 3 \tab 3 \tab 3 \tab 55 \tab - \tab - \tab 14 \tab 2 \tab 2 \tab 4 \tab - \tab 5 \tab 5 \tab 1 \tab 4 \tab 42 \tab 14 \tab 2 \tab 5 \tab 16 \tab 3 \tab 2 \tab 9 \tab 6 \tab 3 \tab 3 \tab 3 \tab 1 \tab 15 \tab 1 \tab 2 \tab 9 \tab 1 \tab 5 \tab 21 \tab 1 \tab 5 \tab 5 \tab 1 \tab 1 \tab 9 \tab 12 \tab 2 \tab 1 \tab 12 \tab 1 \tab 3 \tab 6 \tab 3 \tab 3 \tab 8 \tab 1 \tab 4 \tab 1 \tab 8 \tab 5 \tab 1 \tab - \tab 12 \tab 26 \tab - \tab 5 \tab 6 \tab 1 \tab 12 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 5 \tab - \tab 1 \tab 7 \tab 4 \tab 1 \tab 1 \tab 3 \tab 4 \tab 4 \tab - \tab 22 \tab 9 \tab 1 \tab 6 \tab 2 \tab 1 \tab 2 \tab 35 \tab 17 \tab 12 \tab 3 \tab 1 \tab 1 \tab 8 \tab 2 \tab 2 \tab 2 \tab 1 \tab 3 \tab - \tab 2 \tab 63 \tab 13 \tab 11 \tab 2 \tab 1 \tab 1 \tab - \tab 9 \tab 4 \tab 1 \tab 2 \tab 39 \tab 3 \tab 3 \tab 8 \tab 1 \tab 6 \tab 7 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 9 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 3 \tab - \tab - \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1537 \cr
   1990 \tab 217 \tab 23 \tab 1 \tab 1 \tab 180 \tab 5 \tab 5 \tab 13 \tab 1 \tab 7 \tab 4 \tab 17 \tab 2 \tab 6 \tab 2 \tab 3 \tab 34 \tab 1 \tab 3 \tab 10 \tab 8 \tab 4 \tab 2 \tab 5 \tab 2 \tab - \tab 3 \tab 7 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 9 \tab - \tab 8 \tab 9 \tab 1 \tab - \tab 18 \tab 2 \tab 1 \tab 10 \tab 3 \tab 46 \tab - \tab - \tab 8 \tab 5 \tab 7 \tab - \tab 6 \tab 5 \tab 12 \tab 6 \tab 4 \tab 27 \tab 18 \tab 5 \tab 5 \tab 16 \tab 6 \tab 3 \tab 5 \tab 7 \tab 2 \tab 3 \tab 2 \tab 2 \tab 15 \tab 1 \tab 1 \tab 12 \tab - \tab 1 \tab 18 \tab 1 \tab 4 \tab 4 \tab 5 \tab 6 \tab 14 \tab 13 \tab 4 \tab - \tab 6 \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 5 \tab 5 \tab 2 \tab - \tab 2 \tab 10 \tab 32 \tab 1 \tab 7 \tab - \tab 5 \tab 4 \tab 7 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 10 \tab 1 \tab 14 \tab 2 \tab - \tab 10 \tab 2 \tab 3 \tab 1 \tab 36 \tab 6 \tab 8 \tab 2 \tab 2 \tab 3 \tab 5 \tab 2 \tab 4 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 65 \tab 5 \tab 7 \tab - \tab - \tab 1 \tab 1 \tab 10 \tab 2 \tab 2 \tab - \tab 19 \tab 1 \tab 4 \tab 5 \tab 2 \tab 6 \tab 5 \tab 4 \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 5 \tab 4 \tab 5 \tab - \tab 1 \tab 3 \tab - \tab 4 \tab 6 \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 4 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 49681 \tab 437 \tab 26 \tab 14 \tab 4551 \tab 105 \tab 59 \tab 176 \tab 27 \tab 129 \tab 75 \tab 219 \tab 60 \tab 105 \tab 101 \tab 78 \tab 429 \tab 17 \tab 37 \tab 570 \tab 136 \tab 54 \tab 23 \tab 113 \tab 45 \tab 3 \tab 111 \tab 90 \tab 8 \tab 26 \tab 49 \tab 26 \tab 20 \tab 24 \tab 21 \tab 245 \tab 8 \tab 140 \tab 99 \tab 24 \tab 33 \tab 297 \tab 65 \tab 29 \tab 110 \tab 39 \tab 687 \tab 32 \tab 10 \tab 167 \tab 65 \tab 59 \tab 23 \tab 49 \tab 88 \tab 126 \tab 49 \tab 100 \tab 875 \tab 315 \tab 74 \tab 56 \tab 315 \tab 69 \tab 23 \tab 57 \tab 123 \tab 65 \tab 55 \tab 53 \tab 48 \tab 229 \tab 29 \tab 28 \tab 96 \tab 16 \tab 84 \tab 458 \tab 7 \tab 105 \tab 78 \tab 39 \tab 53 \tab 170 \tab 113 \tab 173 \tab 12 \tab 68 \tab 57 \tab 42 \tab 61 \tab 105 \tab 53 \tab 138 \tab 30 \tab 49 \tab 49 \tab 106 \tab 71 \tab 38 \tab 39 \tab 183 \tab 456 \tab 7 \tab 100 \tab 45 \tab 54 \tab 96 \tab 42 \tab 13 \tab 35 \tab 24 \tab 17 \tab 17 \tab 8 \tab 40 \tab 101 \tab 20 \tab 50 \tab 151 \tab 62 \tab 27 \tab 12 \tab 32 \tab 47 \tab 65 \tab 26 \tab 235 \tab 126 \tab 60 \tab 90 \tab 32 \tab 50 \tab 26 \tab 742 \tab 179 \tab 176 \tab 54 \tab 97 \tab 52 \tab 249 \tab 21 \tab 40 \tab 38 \tab 31 \tab 115 \tab 19 \tab 18 \tab 954 \tab 140 \tab 107 \tab 35 \tab 18 \tab 50 \tab 20 \tab 113 \tab 35 \tab 36 \tab 25 \tab 391 \tab 40 \tab 56 \tab 102 \tab 12 \tab 54 \tab 141 \tab 31 \tab 40 \tab 66 \tab 11 \tab 47 \tab 20 \tab 75 \tab 39 \tab 32 \tab 8 \tab 21 \tab 43 \tab 6 \tab 28 \tab 57 \tab 13 \tab 15 \tab 24 \tab 27 \tab 9 \tab 15 \tab 8 \tab 24 \tab 12 \tab 10 \tab 11 \tab 14 \tab 66 \tab 2 \tab 17 \tab 8 \tab 27 \tab 20 \tab 9 \tab 17 \tab 21 \tab 15 \tab 3 \tab 3 \tab 7 \tab 11 \tab 8 \tab 13 \tab 2 \tab 3 \tab 4 \tab 1 \tab 5 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{17} agricultural production
\item \code{18} agricultural services, except horticultural
\item \code{19} horticultural services
\item \code{27} forestry
\item \code{28} fisheries
\item \code{47} metal mining
\item \code{48} coal mining
\item \code{49} crude petroleum and natural gas extractions
\item \code{57} nonmetallic mining and quarrying, except fuel
\item \code{67} general building contractors
\item \code{68} general contractors, except building
\item \code{69} special trade contractors
\item \code{77} not specified construction
\item \code{107} logging
\item \code{108} sawmills, planing mills, and mill work
\item \code{109} miscellaneous wood products
\item \code{118} furniture fixtures
\item \code{119} glass and glass products
\item \code{127} cement concrete, gypsum, and plaster products
\item \code{128} structural clay products
\item \code{137} pottery and related products
\item \code{138} miscellaneous nonmetallic mineral and stone products
\item \code{139} blast furnaces, steel works, rolling and finishing mills
\item \code{147} other primary iron and steel industries
\item \code{148} primary aluminum industries
\item \code{149} other primary nonferrous industries
\item \code{157} cutlery, hand tools, and other hardware
\item \code{158} fabricated structural metal products
\item \code{159} screw machine products
\item \code{167} metal stamping
\item \code{168} miscellaneous fabricated metal products
\item \code{169} not specified metal industries
\item \code{177} engines and turbines
\item \code{178} farm machinery and equipment
\item \code{179} construction and material handling machines
\item \code{187} metalworking machinery
\item \code{188} office and accounting machines
\item \code{189} electronic computing equipment
\item \code{197} machinery, except electrical, n.e.c.
\item \code{198} not specified machinery
\item \code{199} household appliances
\item \code{207} radio, t.v., and communication equipment
\item \code{208} electrical machinery, equipment, and supplies, n.e.c.
\item \code{209} not specified electrical machinery, equipment, and supplies
\item \code{219} motor vehicles and motor vehicle equipment
\item \code{227} aircraft and parts
\item \code{228} ship and boat building and repairing
\item \code{229} railroad locomotives and equipment
\item \code{237} mobile dwellings and campers
\item \code{238} cycles and miscellaneous transportation equipment
\item \code{239} scientific and controlling instruments
\item \code{247} optical and health services supplies
\item \code{248} photographic equipment and supplies
\item \code{249} watches, clocks, and clockwork-operated devices
\item \code{257} not specified professional equipment
\item \code{258} ordinance
\item \code{259} miscellaneous manufacturing industries
\item \code{268} meat products
\item \code{269} dairy products
\item \code{278} canning and preserving fruits, vegetables, and sea foods
\item \code{279} grain-mill products
\item \code{287} bakery products
\item \code{288} confectionery and related products
\item \code{289} beverage industries
\item \code{297} miscellaneous food preparation and kindred products
\item \code{298} not specified food industries
\item \code{299} tobacco manufactures
\item \code{307} knitting mills
\item \code{308} dyeing and finishing textiles, except wool and knit goods
\item \code{309} floor coverings, except hard surface
\item \code{317} yarn, thread, and fabric mills
\item \code{318} miscellaneous textile mill products
\item \code{319} apparel and accessories
\item \code{327} miscellaneous fabricated textile products
\item \code{328} pulp, paper and paperboard mills
\item \code{329} miscellaneous paper and pulp products
\item \code{337} paperboard containers and boxes
\item \code{338} newspaper publishing and printing
\item \code{339} printing, publishing, and allied industries, except newspapers
\item \code{347} industrial chemicals
\item \code{348} plastics, synthetics and resins, except fibers
\item \code{349} synthetic fibers
\item \code{357} drugs and medicines
\item \code{358} soaps and cosmetics
\item \code{359} paints, varnishes, and related products
\item \code{367} agricultural chemicals
\item \code{368} miscellaneous chemicals
\item \code{369} not specified chemicals and allied products
\item \code{377} petroleum refining
\item \code{378} miscellaneous petroleum and coal products
\item \code{379} rubber products
\item \code{387} miscellaneous plastic products
\item \code{388} tanned, curried, and finished leather
\item \code{389} footwear, except rubber
\item \code{397} leather products, except footwear
\item \code{398} not specified manufacturing industries
\item \code{407} railroads and railway express
\item \code{408} street railways and bus lines
\item \code{409} taxicab service
\item \code{417} trucking service
\item \code{418} warehousing and storage
\item \code{419} water transportation
\item \code{427} air transportation
\item \code{428} pipe lines, except natural gas
\item \code{429} services incidental to transportation
\item \code{447} radio broadcasting and television
\item \code{448} telephone (wire and radio)
\item \code{449} telegraph and miscellaneous communication services
\item \code{467} electric light and power
\item \code{468} electric-gas utilities
\item \code{469} gas and steam supply systems
\item \code{477} water supply
\item \code{478} sanitary services
\item \code{479} other and not specified utilities
\item \code{507} motor vehicles and equipment
\item \code{508} drugs, chemicals, and allied products
\item \code{509} dry goods and apparel
\item \code{527} food and related products
\item \code{528} farm products-raw materials
\item \code{529} electrical goods
\item \code{537} hardware, plumbing, and heating supplies
\item \code{538} not specified electrical and hardware products
\item \code{539} machinery equipment and supplies
\item \code{557} metals and minerals, n.e.c.
\item \code{558} petroleum products
\item \code{559} scrap and waste materials
\item \code{567} alcoholic beverages
\item \code{568} paper and its products
\item \code{569} lumber and construction materials
\item \code{587} wholesalers, n.e.c.
\item \code{588} not specified wholesale trade
\item \code{607} lumber and building material retailing
\item \code{608} hardware and farm equipment stores
\item \code{609} department and mail order establishments
\item \code{617} limited price variety stores
\item \code{618} vending machine operators
\item \code{619} direct selling establishments
\item \code{627} misc. general merchandise stores
\item \code{628} grocery stores
\item \code{629} dairy product stores
\item \code{637} retail bakeries
\item \code{638} food stores, n.e.c.
\item \code{639} motor vehicle dealers
\item \code{647} tire, battery, and accessory dealers
\item \code{648} gasoline service stations
\item \code{649} miscellaneous vehicle dealers
\item \code{657} apparel and accessories stores, except shoe stores
\item \code{658} shoe stores
\item \code{667} furniture and home furnishing stores
\item \code{668} household appliances, tv, and radio stores
\item \code{669} eating and drinking places
\item \code{677} drug stores
\item \code{678} liquor stores
\item \code{679} farm and garden supply stores
\item \code{687} jewelry stores
\item \code{688} fuel and ice dealers
\item \code{689} retail florists
\item \code{697} miscellaneous retail stores
\item \code{698} not specified retail trade
\item \code{707} banking
\item \code{708} credit agencies
\item \code{709} security, commodity brokerage, and investment companies
\item \code{717} insurance
\item \code{718} real estate, incl. real estate-insurance-law offices
\item \code{727} advertising
\item \code{728} services to dwellings and other buildings
\item \code{729} commercial research, development, and testing labs
\item \code{737} employment and temporary help agencies
\item \code{738} business management and consulting services
\item \code{739} computer programming services
\item \code{747} detective and protective services
\item \code{748} business services, n.e.c.
\item \code{749} automobile services, except repair
\item \code{757} automobile repair and related services
\item \code{758} electrical repair shops
\item \code{759} miscellaneous repair services
\item \code{769} private households
\item \code{777} hotels and motels
\item \code{778} lodging places, except hotels and motels
\item \code{779} laundering, cleaning, and other garment services
\item \code{787} beauty shops
\item \code{788} barber shops
\item \code{789} shoe repair shops
\item \code{797} dressmaking shops
\item \code{798} miscellaneous personal services
\item \code{807} theaters and motion pictures
\item \code{808} bowling alleys, billiard and pool parlors
\item \code{809} miscellaneous entertainment and recreation services
\item \code{828} offices of physicians
\item \code{829} offices of dentists
\item \code{837} offices of chiropractors
\item \code{838} hospitals
\item \code{839} convalescent institutions
\item \code{847} offices of health practitioners,n.e.c.
\item \code{848} health services, n.e.c.
\item \code{849} legal services
\item \code{857} elementary and secondary schools
\item \code{858} colleges and universities
\item \code{859} libraries
\item \code{867} educational services, n.e.c.
\item \code{868} not specified educational services
\item \code{869} museums, art galleries, and zoos
\item \code{877} religious organizations
\item \code{878} welfare services
\item \code{879} residential welfare facilities
\item \code{887} nonprofit membership organizations
\item \code{888} engineering and architectural services
\item \code{889} accounting, auditing, and bookkeeping services
\item \code{897} miscellaneous professional and related services
\item \code{907} postal service
\item \code{917} federal public administration
\item \code{927} state public administration
\item \code{937} local public administration
}
}

\keyword{variable}
