% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_07.R
\name{paind80}
\alias{paind80}
\title{Father's industry code (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
paind80
}
\details{
Question 9. Father's occupation, prestige, and industry.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab accounting, auditing, and bookkeeping services \tab administration of economic programs \tab administration of environmental quality and household programs \tab advertising \tab agricultural production, crops \tab agricultural production, livestock \tab agricultural services, except horticultural \tab air transportation \tab aircraft and parts \tab apparel and accessories, except knit \tab apparel and accessory stores, except shoe \tab apparel, fabrics, and notions \tab auto and home supply stores \tab automotive repair shops \tab automotive services, except repair \tab bakery products \tab banking \tab barber shops \tab beverage industries \tab blast furnaces, steelworks, rolling and finishing mills \tab bus service and urban transit \tab business management and consulting services \tab business services, n.e.c. \tab canned and preserved fruits and vegetables \tab cement, concrete, gypsum, and plaster products \tab coal mining \tab colleges and universities \tab commercial research, development, and testing labs \tab construction \tab construction and material handling machines \tab credit agencies, n.e.c. \tab crude petroleum and natural gas extraction \tab cutlery, hand tools, and other hardware \tab cycles and miscellaneous transportation equipment \tab dairy products \tab department stores \tab detective and protective services \tab don't know \tab drug stores \tab drugs \tab drugs, chemicals, and allied products \tab eating and drinking places \tab educational services, n.e.c. \tab electric and gas, and other combinations \tab electric light and power \tab electrical goods \tab electrical machinery, equipment, and supplies, n.e.c. \tab electrical repair shops \tab electronic computing equipment \tab elementary and secondary schools \tab engineering, architectural, and surveying services \tab engines and turbines \tab executive and legislative offices \tab fabricated structural metal products \tab farm products-raw materials \tab farm supplies \tab fishing, hunting, and trapping, etc. \tab food stores, n.e.c. \tab footwear, except rubber and plastic \tab forestry \tab fuel and ice dealers \tab funeral service and crematories \tab furniture and fixtures \tab furniture and home furnishings \tab furniture and home furnishings stores \tab gas and steam supply systems \tab gasoline service stations \tab general government, n.e.c. \tab glass and glass products \tab grain mill products \tab groceries and related products \tab grocery stores \tab guided missiles, space vehicles, and parts \tab hardware stores \tab hardware, plumbing and heating supplies \tab health services, n.e.c. \tab horticultural services \tab hospitals \tab hotels and motels \tab household appliances \tab household appliances, tv, and radio stores \tab insurance \tab iron and steel foundries \tab jewelry stores \tab job training and vocational rehabilitation services \tab justice, public order, and safety \tab laundry, cleaning, and garment services \tab leather products, except footwear \tab legal services \tab liquor stores \tab lodging places, except hotels and motels \tab logging \tab lumber and building material retailing \tab lumber and construction materials \tab machinery, equipment, and supplies \tab machinery, except electrical, n.e.c. \tab meat products \tab membership organizations \tab metal forgings and stampings \tab metal mining \tab metalworking machinery \tab miscellaneous entertainment and recreation services \tab miscellaneous fabricated metal products \tab miscellaneous fabricated textile products \tab miscellaneous food preparations and kindred products \tab miscellaneous general merchandise stores \tab miscellaneous manufacturing industries \tab miscellaneous nonmetallic mineral and stone products \tab miscellaneous paper and pulp products \tab miscellaneous personal services \tab miscellaneous petroleum and coal products \tab miscellaneous plastics products \tab miscellaneous professional and related services \tab miscellaneous repair services \tab miscellaneous retail stores \tab miscellaneous textile mill products \tab miscellaneous vehicle dealers \tab miscellaneous wholesale, durable goods \tab miscellaneous wholesale, nondurable goods \tab miscellaneous wood products \tab motor vehicle dealers \tab motor vehicles and equipment \tab motor vehicles and motor vehicle equipment \tab museums, art galleries and zoos \tab national security and international affairs \tab newspaper publishing and printing \tab no answer \tab noncommercial educational and scientific research \tab not specified electrical machinery, equipment, and supplies \tab not specified food industries \tab not specified machinery \tab not specified manufacturing industries \tab not specified metal industries \tab not specified retail trade \tab not specified utilities \tab not specified wholesale trade \tab offices of dentists \tab offices of physicians \tab optical and health services supplies \tab ordnance \tab other primary metal industries \tab other rubber products, and plastics footwear and belting \tab paper and paper products \tab paperboard containers and boxes \tab personnel supply services \tab petroleum and coal products \tab petroleum products \tab petroleum refining \tab photographic equipment and supplies \tab plastics, synthetics, and resins \tab pottery and related products \tab primary aluminum industries \tab printing, publishing, and allied industries, except newspapers \tab private households \tab pulp, paper, and paperboard mills \tab radio and television broadcasting \tab radio, tv, and communication equipment \tab railroad locomotives and equipment \tab railroads \tab real estate, including real estate-insurance-law offices \tab religious organizations \tab retail bakeries \tab retail nurseries and garden stores \tab sanitary services \tab sawmills, planning mills, and millwork \tab scientific and controlling instruments \tab screw machine products \tab security, commodity brokerage, and investment companies \tab services incidental to transportation \tab services to dwellings and other buildings \tab sewing, needlework, and piece goods stores \tab ship and boat building and repairing \tab shoe repair shops \tab shoe stores \tab structural clay products \tab sugar and confectionery products \tab taxicab service \tab telegraph and miscellaneous communication service \tab telephone (wire and radio) \tab theaters and motion pictures \tab tires and inner tubes \tab trucking service \tab u.s. postal service \tab variety stores \tab vending machine operators \tab warehousing and storage \tab watches, clocks, and clockwork operated devices \tab water supply and irrigation \tab water transportation \tab wood buildings and mobile homes \tab yarn, thread, and fabric mills \tab administration of human resources programs \tab book and stationery stores \tab bowling alleys, billiard and pool parlors \tab computer and data processing services \tab direct selling establishments \tab farm machinery and equipment \tab floor coverings, except hard surface \tab industrial and miscellaneous chemicals \tab knitting mills \tab leather tanning and finishing \tab metals and minerals, except petroleum \tab nonmetallic mining and quarrying, except fuel \tab office and accounting machines \tab offices of chiropractors \tab pipe lines, except natural gas \tab public finance, taxation, and monetary policy \tab scrap and waste materials \tab soaps and cosmetics \tab social services, n.e.c. \tab sporting goods, bicycles, and hobby stores \tab tobacco manufactures \tab toys, amusement, and sporting goods \tab alcoholic beverages \tab mail order houses \tab mobile home dealers \tab offices of optometrists \tab retail florists \tab savings and loan associations \tab sporting goods, toys, and hobby goods \tab agricultural chemicals \tab beauty shops \tab business, trade, and vocational schools \tab libraries \tab nursing and personal care facilities \tab dyeing and finishing textiles, except wool and knit goods \tab not specified electrical and hardware products \tab residential care facilities, without nursing \tab offices of health practitioners, n.e.c. \tab dressmaking shops \tab dairy products stores \tab child day care services \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 207 \tab 2 \tab 3 \tab 2 \tab 3 \tab 157 \tab 63 \tab 5 \tab 6 \tab 7 \tab 2 \tab 2 \tab 2 \tab 3 \tab 13 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 15 \tab 11 \tab 1 \tab 4 \tab 1 \tab 5 \tab 15 \tab 15 \tab 1 \tab 124 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 6 \tab 1 \tab 7 \tab 2 \tab 1 \tab 2 \tab 13 \tab 1 \tab 5 \tab 8 \tab 3 \tab 6 \tab 6 \tab 1 \tab 26 \tab 3 \tab 2 \tab 4 \tab 6 \tab 4 \tab 4 \tab 1 \tab 2 \tab 2 \tab 2 \tab 3 \tab 1 \tab 8 \tab 2 \tab 3 \tab 4 \tab 7 \tab 67 \tab 1 \tab 2 \tab 9 \tab 19 \tab 4 \tab 5 \tab 5 \tab 1 \tab 4 \tab 12 \tab 5 \tab 1 \tab 3 \tab 10 \tab 10 \tab 2 \tab 1 \tab 15 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab 4 \tab 6 \tab 4 \tab 4 \tab 8 \tab 7 \tab 4 \tab 2 \tab 1 \tab 2 \tab 3 \tab 4 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 7 \tab 5 \tab 7 \tab 1 \tab 2 \tab 1 \tab 2 \tab 15 \tab 3 \tab 18 \tab 1 \tab 15 \tab 7 \tab 12 \tab 1 \tab 4 \tab 1 \tab 3 \tab 20 \tab 2 \tab 3 \tab 2 \tab 3 \tab 3 \tab 4 \tab 1 \tab 3 \tab 5 \tab 3 \tab 1 \tab 4 \tab 1 \tab 11 \tab 3 \tab 7 \tab 2 \tab 2 \tab 3 \tab 2 \tab 5 \tab 2 \tab 5 \tab 5 \tab 2 \tab 1 \tab 31 \tab 8 \tab 13 \tab 4 \tab 2 \tab 9 \tab 12 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2 \tab 4 \tab 2 \tab 1 \tab 7 \tab 2 \tab 5 \tab 2 \tab 2 \tab 27 \tab 9 \tab 5 \tab 1 \tab 9 \tab 1 \tab 1 \tab 2 \tab 1 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 225 \tab 4 \tab 1 \tab 1 \tab - \tab 120 \tab 116 \tab 10 \tab 4 \tab 6 \tab 2 \tab 1 \tab - \tab 4 \tab 19 \tab 1 \tab 4 \tab 6 \tab 4 \tab 5 \tab 30 \tab 13 \tab - \tab 3 \tab 3 \tab 2 \tab 20 \tab 14 \tab 1 \tab 125 \tab 5 \tab 1 \tab 6 \tab - \tab 1 \tab 3 \tab 5 \tab 1 \tab 5 \tab 4 \tab 1 \tab 1 \tab 18 \tab 1 \tab 3 \tab 9 \tab 1 \tab 9 \tab 3 \tab 3 \tab 24 \tab 10 \tab 5 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 5 \tab 2 \tab 3 \tab 1 \tab 5 \tab 14 \tab 3 \tab 4 \tab 16 \tab 14 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 5 \tab 2 \tab 3 \tab 1 \tab 15 \tab 6 \tab 1 \tab - \tab 20 \tab 11 \tab 1 \tab 6 \tab 1 \tab - \tab 7 \tab 4 \tab 2 \tab 3 \tab 11 \tab 13 \tab 3 \tab 7 \tab 9 \tab 8 \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab 2 \tab 4 \tab 1 \tab - \tab 2 \tab - \tab 6 \tab 3 \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 13 \tab - \tab 22 \tab - \tab 40 \tab 7 \tab 18 \tab 1 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 5 \tab 3 \tab 3 \tab 3 \tab 1 \tab 1 \tab 4 \tab - \tab 7 \tab 2 \tab 6 \tab - \tab 4 \tab - \tab 3 \tab 8 \tab 1 \tab 4 \tab 1 \tab 4 \tab 4 \tab 30 \tab 16 \tab 13 \tab 1 \tab - \tab 2 \tab 10 \tab 1 \tab 1 \tab 2 \tab 6 \tab 3 \tab - \tab 4 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 6 \tab 4 \tab 2 \tab 37 \tab 20 \tab - \tab - \tab - \tab 2 \tab 3 \tab 7 \tab - \tab 11 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 7 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 217 \tab 1 \tab 3 \tab 3 \tab 4 \tab 107 \tab 75 \tab 4 \tab 6 \tab 13 \tab 6 \tab 4 \tab 2 \tab 2 \tab 17 \tab - \tab 3 \tab 4 \tab 3 \tab 4 \tab 23 \tab 9 \tab 1 \tab - \tab 3 \tab 4 \tab 12 \tab 8 \tab - \tab 115 \tab 4 \tab 1 \tab 4 \tab 1 \tab - \tab 5 \tab 5 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 11 \tab 1 \tab - \tab 10 \tab 2 \tab 9 \tab - \tab 2 \tab 18 \tab 2 \tab 1 \tab 3 \tab 8 \tab 4 \tab 1 \tab - \tab 4 \tab 5 \tab - \tab 6 \tab 1 \tab 11 \tab 2 \tab 6 \tab 7 \tab 7 \tab 12 \tab 5 \tab 4 \tab 6 \tab 14 \tab - \tab 4 \tab 4 \tab 3 \tab 6 \tab 7 \tab 2 \tab 3 \tab 1 \tab 13 \tab 13 \tab - \tab - \tab 14 \tab 11 \tab - \tab 1 \tab - \tab 1 \tab 5 \tab 8 \tab 3 \tab 6 \tab 11 \tab 15 \tab 3 \tab 1 \tab 4 \tab 10 \tab 3 \tab 4 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 2 \tab 4 \tab 7 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 9 \tab 2 \tab 31 \tab 1 \tab 25 \tab 7 \tab 22 \tab - \tab 3 \tab - \tab 4 \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 2 \tab 5 \tab 1 \tab 4 \tab 2 \tab - \tab - \tab 1 \tab - \tab 9 \tab 3 \tab 8 \tab 1 \tab 3 \tab - \tab 6 \tab 2 \tab - \tab 8 \tab 2 \tab 3 \tab 1 \tab 35 \tab 6 \tab 8 \tab 2 \tab - \tab 3 \tab 4 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab - \tab 6 \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 11 \tab 1 \tab 2 \tab 18 \tab 15 \tab - \tab - \tab 2 \tab 2 \tab 4 \tab 3 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 5 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 220 \tab 2 \tab 5 \tab 1 \tab 3 \tab 131 \tab 77 \tab 9 \tab 4 \tab 11 \tab 5 \tab 8 \tab 3 \tab 5 \tab 12 \tab 1 \tab 7 \tab 4 \tab 5 \tab 4 \tab 25 \tab 6 \tab 3 \tab 2 \tab 3 \tab 3 \tab 25 \tab 9 \tab 2 \tab 145 \tab 4 \tab 1 \tab 8 \tab - \tab - \tab 4 \tab 8 \tab 2 \tab 15 \tab 2 \tab 3 \tab - \tab 14 \tab - \tab 3 \tab 9 \tab - \tab 7 \tab 2 \tab 2 \tab 24 \tab 7 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 8 \tab 2 \tab 6 \tab 4 \tab 4 \tab 4 \tab 10 \tab 19 \tab 4 \tab 3 \tab 3 \tab 2 \tab 6 \tab 11 \tab 3 \tab 3 \tab 4 \tab 15 \tab 14 \tab - \tab 1 \tab 7 \tab 6 \tab - \tab 3 \tab - \tab 1 \tab 3 \tab 3 \tab 3 \tab 3 \tab 9 \tab 10 \tab 2 \tab 3 \tab 2 \tab 8 \tab 4 \tab 2 \tab - \tab 3 \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 9 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 3 \tab 11 \tab 2 \tab 33 \tab 1 \tab 40 \tab 7 \tab 28 \tab - \tab 2 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 4 \tab 4 \tab 9 \tab 1 \tab 5 \tab 2 \tab 2 \tab 2 \tab 3 \tab - \tab 8 \tab 3 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 11 \tab - \tab 8 \tab 1 \tab - \tab 3 \tab 35 \tab 8 \tab 11 \tab 1 \tab - \tab 6 \tab 10 \tab 2 \tab 1 \tab 2 \tab - \tab 4 \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 10 \tab 3 \tab 3 \tab 27 \tab 21 \tab 1 \tab 1 \tab 3 \tab - \tab 5 \tab 1 \tab - \tab 4 \tab 4 \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 197 \tab - \tab 3 \tab 4 \tab 3 \tab 177 \tab 57 \tab 10 \tab 6 \tab 5 \tab 8 \tab 5 \tab 2 \tab 4 \tab 24 \tab 2 \tab 6 \tab 11 \tab 3 \tab 1 \tab 3 \tab 8 \tab - \tab 3 \tab 1 \tab 4 \tab 9 \tab 12 \tab - \tab 137 \tab 4 \tab - \tab 6 \tab 3 \tab 1 \tab 8 \tab 5 \tab - \tab - \tab 2 \tab 5 \tab 1 \tab 19 \tab 1 \tab 1 \tab 15 \tab 1 \tab 6 \tab 2 \tab 8 \tab 24 \tab 7 \tab 2 \tab 3 \tab 2 \tab 2 \tab - \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 9 \tab - \tab 4 \tab - \tab 9 \tab 12 \tab 5 \tab 3 \tab 8 \tab 18 \tab 5 \tab 3 \tab 5 \tab 1 \tab 9 \tab 11 \tab 5 \tab 1 \tab 2 \tab 16 \tab 21 \tab 5 \tab - \tab 32 \tab 5 \tab - \tab 9 \tab 2 \tab - \tab 8 \tab 5 \tab 2 \tab 6 \tab 16 \tab 5 \tab - \tab 2 \tab 2 \tab 9 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 8 \tab 2 \tab 7 \tab 1 \tab 1 \tab - \tab - \tab 8 \tab 3 \tab 4 \tab 2 \tab 1 \tab - \tab 3 \tab 15 \tab 1 \tab 33 \tab 1 \tab 35 \tab 2 \tab 41 \tab - \tab 1 \tab 2 \tab 2 \tab 7 \tab 2 \tab 3 \tab 2 \tab 2 \tab 4 \tab 7 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 4 \tab - \tab 13 \tab 2 \tab 5 \tab - \tab 4 \tab - \tab 3 \tab 15 \tab 2 \tab 6 \tab 2 \tab 4 \tab 4 \tab 32 \tab 8 \tab 11 \tab - \tab 1 \tab 6 \tab 15 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 10 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab 2 \tab 13 \tab 3 \tab 4 \tab 28 \tab 17 \tab 1 \tab 1 \tab 1 \tab - \tab 5 \tab 5 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 4 \tab - \tab 5 \tab 3 \tab - \tab - \tab 6 \tab 1 \tab 5 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 508 \tab 9 \tab 9 \tab 2 \tab 1 \tab 217 \tab 131 \tab 6 \tab 11 \tab 28 \tab 11 \tab 11 \tab 1 \tab 3 \tab 30 \tab 4 \tab 12 \tab 18 \tab 10 \tab 7 \tab 15 \tab 17 \tab 1 \tab 4 \tab 4 \tab 5 \tab 16 \tab 20 \tab 3 \tab 232 \tab 11 \tab - \tab 12 \tab 2 \tab - \tab 16 \tab 9 \tab 1 \tab 2 \tab 6 \tab 6 \tab 2 \tab 30 \tab 3 \tab 3 \tab 21 \tab 4 \tab 26 \tab 4 \tab 10 \tab 50 \tab 11 \tab 1 \tab 3 \tab 13 \tab 1 \tab 5 \tab 6 \tab 4 \tab 5 \tab 1 \tab 6 \tab 2 \tab 11 \tab 2 \tab 5 \tab 4 \tab 15 \tab 13 \tab 10 \tab 9 \tab 10 \tab 32 \tab 8 \tab 3 \tab 3 \tab 5 \tab 6 \tab 17 \tab 10 \tab 11 \tab 8 \tab 34 \tab 40 \tab 4 \tab - \tab 47 \tab 6 \tab - \tab 16 \tab 1 \tab - \tab 12 \tab 11 \tab 2 \tab 8 \tab 26 \tab 12 \tab 13 \tab 7 \tab 8 \tab 15 \tab 15 \tab 12 \tab 3 \tab 5 \tab 2 \tab 12 \tab 7 \tab 5 \tab - \tab - \tab 6 \tab 1 \tab 9 \tab 7 \tab 14 \tab 1 \tab 1 \tab 1 \tab 5 \tab 20 \tab 6 \tab 44 \tab 2 \tab 73 \tab 11 \tab 36 \tab 2 \tab 6 \tab 1 \tab 1 \tab 13 \tab - \tab 2 \tab 2 \tab 1 \tab 6 \tab 7 \tab 4 \tab 5 \tab 5 \tab 6 \tab 1 \tab 2 \tab 1 \tab 21 \tab 3 \tab 23 \tab 3 \tab 7 \tab 3 \tab 4 \tab 16 \tab 3 \tab 15 \tab 2 \tab 3 \tab 2 \tab 84 \tab 17 \tab 20 \tab 3 \tab 2 \tab 9 \tab 26 \tab 4 \tab - \tab 4 \tab 2 \tab 5 \tab - \tab 17 \tab 3 \tab 2 \tab 2 \tab 6 \tab 8 \tab 1 \tab 24 \tab 7 \tab 5 \tab 50 \tab 28 \tab - \tab 3 \tab 3 \tab - \tab 6 \tab 11 \tab - \tab 18 \tab 4 \tab - \tab - \tab 4 \tab 9 \tab 11 \tab 1 \tab 4 \tab 4 \tab 1 \tab - \tab 17 \tab 7 \tab 4 \tab 1 \tab 4 \tab 2 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 502 \tab 10 \tab 11 \tab 12 \tab 6 \tab 139 \tab 158 \tab 4 \tab 16 \tab 17 \tab 10 \tab 7 \tab 1 \tab 6 \tab 26 \tab 3 \tab 6 \tab 16 \tab 10 \tab 6 \tab 26 \tab 11 \tab - \tab 9 \tab 5 \tab 12 \tab 19 \tab 20 \tab - \tab 276 \tab 6 \tab - \tab 19 \tab 2 \tab - \tab 7 \tab 13 \tab 1 \tab - \tab 7 \tab 2 \tab - \tab 22 \tab 1 \tab 3 \tab 29 \tab 4 \tab 12 \tab 8 \tab 8 \tab 54 \tab 17 \tab - \tab 2 \tab 5 \tab 2 \tab 4 \tab 7 \tab 5 \tab 7 \tab 3 \tab 5 \tab 4 \tab 13 \tab 1 \tab 14 \tab 6 \tab 13 \tab 13 \tab 6 \tab 6 \tab 12 \tab 38 \tab 12 \tab 5 \tab 3 \tab 1 \tab 6 \tab 20 \tab 4 \tab 9 \tab 6 \tab 23 \tab 26 \tab 3 \tab - \tab 44 \tab 7 \tab 2 \tab 15 \tab 2 \tab - \tab 6 \tab 10 \tab 3 \tab 10 \tab 21 \tab 8 \tab 6 \tab 2 \tab 4 \tab 7 \tab 5 \tab 8 \tab 5 \tab 4 \tab 4 \tab 11 \tab 8 \tab 1 \tab 2 \tab 1 \tab 5 \tab 2 \tab 8 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 24 \tab 5 \tab 61 \tab - \tab 95 \tab 5 \tab 64 \tab 1 \tab 6 \tab 4 \tab 3 \tab 22 \tab 5 \tab 3 \tab 3 \tab 10 \tab 4 \tab 11 \tab 4 \tab 9 \tab 8 \tab - \tab - \tab 6 \tab 1 \tab 16 \tab 7 \tab 8 \tab 5 \tab 4 \tab 2 \tab 8 \tab 14 \tab 5 \tab 9 \tab 4 \tab 3 \tab 2 \tab 64 \tab 18 \tab 24 \tab 5 \tab - \tab 11 \tab 25 \tab 3 \tab - \tab 5 \tab 2 \tab 5 \tab 1 \tab 9 \tab - \tab - \tab - \tab 5 \tab 4 \tab 2 \tab 16 \tab 5 \tab 5 \tab 61 \tab 24 \tab - \tab 2 \tab 3 \tab 1 \tab 4 \tab 7 \tab - \tab 19 \tab 5 \tab 2 \tab 1 \tab 2 \tab - \tab 10 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 6 \tab 1 \tab - \tab - \tab 4 \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 503 \tab 6 \tab 4 \tab 6 \tab 2 \tab 155 \tab 106 \tab 9 \tab 5 \tab 21 \tab 6 \tab 5 \tab 3 \tab 7 \tab 33 \tab 3 \tab 8 \tab 16 \tab 4 \tab 5 \tab 34 \tab 10 \tab 1 \tab 5 \tab 2 \tab 7 \tab 24 \tab 31 \tab 2 \tab 271 \tab 10 \tab 1 \tab 12 \tab 3 \tab 1 \tab 17 \tab 9 \tab 5 \tab - \tab 5 \tab 6 \tab 4 \tab 35 \tab 1 \tab 3 \tab 15 \tab 2 \tab 12 \tab 7 \tab 12 \tab 54 \tab 8 \tab 2 \tab 3 \tab 9 \tab 1 \tab 3 \tab 5 \tab 11 \tab 4 \tab 4 \tab 1 \tab 2 \tab 18 \tab 1 \tab 8 \tab 3 \tab 14 \tab 45 \tab 2 \tab 2 \tab 12 \tab 31 \tab 2 \tab 2 \tab 5 \tab 1 \tab 7 \tab 18 \tab 9 \tab 7 \tab 6 \tab 31 \tab 11 \tab 2 \tab 1 \tab 44 \tab 7 \tab 1 \tab 12 \tab 2 \tab - \tab 9 \tab 4 \tab 9 \tab 10 \tab 20 \tab 11 \tab 4 \tab 2 \tab 8 \tab 3 \tab 5 \tab 5 \tab 1 \tab 7 \tab 2 \tab 9 \tab 3 \tab - \tab 1 \tab 1 \tab 4 \tab 2 \tab 12 \tab 3 \tab - \tab - \tab 3 \tab - \tab 2 \tab 23 \tab 4 \tab 55 \tab - \tab 90 \tab 4 \tab 119 \tab 3 \tab 3 \tab 1 \tab 2 \tab 25 \tab 1 \tab 9 \tab 5 \tab 1 \tab 5 \tab 10 \tab 7 \tab 5 \tab 5 \tab 2 \tab 3 \tab 4 \tab - \tab 14 \tab 9 \tab 15 \tab 1 \tab - \tab 1 \tab 6 \tab 19 \tab - \tab 10 \tab 2 \tab 8 \tab 1 \tab 56 \tab 15 \tab 20 \tab 3 \tab - \tab 11 \tab 18 \tab - \tab 2 \tab 6 \tab 4 \tab 3 \tab 1 \tab 10 \tab 5 \tab 4 \tab - \tab 5 \tab 3 \tab 1 \tab 16 \tab 4 \tab 2 \tab 37 \tab 26 \tab - \tab 1 \tab 3 \tab - \tab 5 \tab 2 \tab - \tab 18 \tab 5 \tab - \tab - \tab 9 \tab - \tab 9 \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab 7 \tab 1 \tab - \tab 1 \tab 4 \tab 3 \tab 2 \tab 3 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 5 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 518 \tab 10 \tab 2 \tab 6 \tab 1 \tab 144 \tab 161 \tab 3 \tab 18 \tab 14 \tab 21 \tab 6 \tab 1 \tab 4 \tab 32 \tab 2 \tab 11 \tab 12 \tab 5 \tab 11 \tab 45 \tab 10 \tab 2 \tab 1 \tab 3 \tab 6 \tab 23 \tab 22 \tab - \tab 252 \tab 6 \tab 5 \tab 19 \tab 4 \tab - \tab 12 \tab 10 \tab 6 \tab - \tab 5 \tab 7 \tab 2 \tab 24 \tab - \tab 6 \tab 26 \tab 5 \tab 18 \tab 2 \tab 5 \tab 43 \tab 10 \tab - \tab 2 \tab 2 \tab 3 \tab 1 \tab 6 \tab 5 \tab 6 \tab 2 \tab - \tab 1 \tab 9 \tab 1 \tab 5 \tab 5 \tab 15 \tab 77 \tab 8 \tab 6 \tab 11 \tab 26 \tab 8 \tab 6 \tab 4 \tab 2 \tab 7 \tab 18 \tab 7 \tab 5 \tab 11 \tab 27 \tab 10 \tab 4 \tab - \tab 37 \tab 9 \tab 1 \tab 11 \tab 1 \tab - \tab 11 \tab 6 \tab 7 \tab 15 \tab 19 \tab 10 \tab 1 \tab 1 \tab - \tab 7 \tab 8 \tab 3 \tab 2 \tab 3 \tab - \tab 7 \tab 4 \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 7 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 26 \tab 2 \tab 59 \tab - \tab 49 \tab 9 \tab 102 \tab 3 \tab 4 \tab 1 \tab - \tab 11 \tab - \tab - \tab 2 \tab 1 \tab 4 \tab 6 \tab 4 \tab 7 \tab 10 \tab 1 \tab 2 \tab 3 \tab - \tab 22 \tab 10 \tab 23 \tab 4 \tab 1 \tab - \tab 6 \tab 18 \tab 4 \tab 19 \tab 4 \tab 3 \tab 1 \tab 51 \tab 22 \tab 23 \tab 2 \tab - \tab 8 \tab 12 \tab 2 \tab 2 \tab 7 \tab 3 \tab 2 \tab 2 \tab 10 \tab - \tab 2 \tab 2 \tab 3 \tab 4 \tab - \tab 12 \tab 6 \tab 6 \tab 52 \tab 22 \tab - \tab - \tab 8 \tab - \tab 6 \tab 9 \tab 1 \tab 18 \tab 1 \tab 2 \tab - \tab 7 \tab 1 \tab 4 \tab 3 \tab - \tab 1 \tab 1 \tab 2 \tab 5 \tab - \tab - \tab - \tab 4 \tab 3 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 2817 \cr
   2002 \tab 526 \tab 10 \tab 6 \tab 8 \tab 2 \tab 115 \tab 145 \tab 1 \tab 8 \tab 15 \tab 5 \tab 5 \tab - \tab 4 \tab 38 \tab 1 \tab 9 \tab 10 \tab 5 \tab 10 \tab 50 \tab 9 \tab 1 \tab 5 \tab 5 \tab 4 \tab 15 \tab 24 \tab - \tab 248 \tab 3 \tab 2 \tab 14 \tab 4 \tab 1 \tab 8 \tab 10 \tab 2 \tab 7 \tab 6 \tab 3 \tab 1 \tab 35 \tab 2 \tab - \tab 30 \tab 1 \tab 9 \tab 6 \tab 6 \tab 53 \tab 10 \tab 5 \tab 1 \tab 4 \tab 4 \tab 2 \tab 3 \tab 3 \tab 5 \tab - \tab 1 \tab 1 \tab 17 \tab - \tab 10 \tab 7 \tab 13 \tab 33 \tab 5 \tab 3 \tab 13 \tab 26 \tab 10 \tab 2 \tab 3 \tab 6 \tab 8 \tab 21 \tab 9 \tab 5 \tab 2 \tab 19 \tab 5 \tab - \tab 1 \tab 50 \tab 4 \tab 4 \tab 11 \tab 2 \tab - \tab 8 \tab 13 \tab 3 \tab 11 \tab 16 \tab 8 \tab 4 \tab 2 \tab 5 \tab 7 \tab 6 \tab 3 \tab - \tab 3 \tab 2 \tab 8 \tab 3 \tab - \tab 2 \tab 1 \tab 4 \tab 3 \tab 9 \tab 4 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 16 \tab - \tab 54 \tab 1 \tab 85 \tab 12 \tab 73 \tab 3 \tab 2 \tab - \tab 1 \tab 14 \tab 3 \tab 4 \tab 1 \tab 2 \tab 5 \tab 8 \tab 3 \tab 2 \tab 1 \tab 3 \tab 3 \tab 4 \tab 1 \tab 20 \tab 6 \tab 21 \tab 4 \tab 9 \tab - \tab 12 \tab 14 \tab 1 \tab 19 \tab 4 \tab 5 \tab 1 \tab 45 \tab 16 \tab 16 \tab 3 \tab - \tab 7 \tab 18 \tab 1 \tab - \tab 9 \tab 4 \tab 10 \tab 2 \tab 12 \tab 1 \tab 1 \tab 2 \tab 3 \tab 6 \tab - \tab 22 \tab 3 \tab 6 \tab 65 \tab 30 \tab 1 \tab 2 \tab 6 \tab - \tab 8 \tab 9 \tab 1 \tab 31 \tab 5 \tab 3 \tab 1 \tab 9 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 5 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab 6 \tab 2 \tab 4 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2765 \cr
   2004 \tab 514 \tab 8 \tab 9 \tab 7 \tab 4 \tab 122 \tab 115 \tab 1 \tab 12 \tab 21 \tab 6 \tab 5 \tab 3 \tab 6 \tab 37 \tab 2 \tab 8 \tab 16 \tab 3 \tab 6 \tab 42 \tab 12 \tab 2 \tab 8 \tab 2 \tab 4 \tab 25 \tab 34 \tab 1 \tab 273 \tab 7 \tab 2 \tab 12 \tab 1 \tab - \tab 9 \tab 11 \tab 2 \tab - \tab 3 \tab 6 \tab 2 \tab 19 \tab - \tab 2 \tab 22 \tab 1 \tab 10 \tab 4 \tab 6 \tab 79 \tab 19 \tab 3 \tab 2 \tab 10 \tab 2 \tab 3 \tab 8 \tab 4 \tab 3 \tab 3 \tab 1 \tab 5 \tab 13 \tab 1 \tab 4 \tab 7 \tab 11 \tab 29 \tab 5 \tab 1 \tab 9 \tab 29 \tab 7 \tab 7 \tab 1 \tab 2 \tab 8 \tab 23 \tab 5 \tab 1 \tab 8 \tab 25 \tab 6 \tab 1 \tab - \tab 35 \tab 5 \tab 1 \tab 15 \tab 1 \tab 4 \tab 16 \tab 6 \tab 6 \tab 12 \tab 27 \tab 12 \tab 6 \tab 2 \tab 2 \tab 10 \tab 7 \tab 5 \tab 3 \tab 4 \tab 2 \tab 7 \tab 5 \tab 2 \tab - \tab 1 \tab 7 \tab 1 \tab 11 \tab 8 \tab - \tab - \tab - \tab 4 \tab 3 \tab 30 \tab - \tab 52 \tab - \tab 97 \tab 11 \tab 108 \tab 5 \tab 4 \tab 2 \tab 2 \tab 21 \tab 2 \tab 5 \tab 1 \tab 3 \tab 5 \tab 12 \tab 1 \tab 3 \tab 3 \tab 4 \tab - \tab 3 \tab - \tab 20 \tab 4 \tab 13 \tab 4 \tab 4 \tab 4 \tab 8 \tab 18 \tab 2 \tab 19 \tab 1 \tab 2 \tab 1 \tab 30 \tab 17 \tab 15 \tab 4 \tab 1 \tab 13 \tab 17 \tab 2 \tab 1 \tab 2 \tab 1 \tab 8 \tab - \tab 11 \tab 1 \tab - \tab 2 \tab 3 \tab 4 \tab 2 \tab 21 \tab 2 \tab 3 \tab 53 \tab 24 \tab - \tab 2 \tab 4 \tab - \tab 6 \tab 4 \tab 2 \tab 13 \tab 6 \tab 1 \tab 2 \tab 10 \tab 3 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 8 \tab 2 \tab 3 \tab 4 \tab - \tab 4 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2049 \tab 2 \tab 13 \tab 7 \tab 1 \tab 173 \tab 90 \tab 5 \tab 9 \tab 8 \tab 14 \tab 8 \tab - \tab 12 \tab 34 \tab - \tab 5 \tab 17 \tab 4 \tab 9 \tab 42 \tab 16 \tab 1 \tab 8 \tab 2 \tab 8 \tab 24 \tab 27 \tab 3 \tab 288 \tab 9 \tab 4 \tab 7 \tab 2 \tab 2 \tab 9 \tab 6 \tab 3 \tab - \tab 7 \tab 4 \tab - \tab 35 \tab - \tab 3 \tab 31 \tab 4 \tab 13 \tab 8 \tab 7 \tab 64 \tab 19 \tab 2 \tab 8 \tab 7 \tab 3 \tab 3 \tab 7 \tab 8 \tab 4 \tab 2 \tab 1 \tab 1 \tab 25 \tab 1 \tab 10 \tab 5 \tab 15 \tab 23 \tab 6 \tab 2 \tab 14 \tab 28 \tab 11 \tab 4 \tab 4 \tab 4 \tab 12 \tab 29 \tab 6 \tab 4 \tab 4 \tab 35 \tab 6 \tab 3 \tab - \tab 48 \tab 8 \tab 1 \tab 16 \tab 2 \tab - \tab 25 \tab 9 \tab 5 \tab 9 \tab 19 \tab 12 \tab 2 \tab 1 \tab 3 \tab 15 \tab 9 \tab 4 \tab 1 \tab 6 \tab - \tab 8 \tab 2 \tab 6 \tab - \tab - \tab 9 \tab - \tab 10 \tab 9 \tab 2 \tab 3 \tab - \tab 1 \tab 5 \tab 26 \tab 2 \tab 52 \tab 3 \tab 80 \tab 10 \tab 108 \tab 6 \tab 4 \tab 5 \tab 4 \tab 12 \tab - \tab 6 \tab 4 \tab - \tab 6 \tab 15 \tab 2 \tab 2 \tab 7 \tab 3 \tab 1 \tab 1 \tab - \tab 22 \tab 10 \tab 13 \tab 5 \tab 5 \tab 1 \tab 8 \tab 10 \tab 4 \tab 12 \tab 1 \tab 1 \tab 3 \tab 41 \tab 14 \tab 22 \tab 3 \tab 1 \tab 10 \tab 20 \tab - \tab 1 \tab 7 \tab 4 \tab 5 \tab 1 \tab 10 \tab 1 \tab - \tab 3 \tab 6 \tab 2 \tab 4 \tab 21 \tab 6 \tab 8 \tab 60 \tab 33 \tab - \tab - \tab 7 \tab 1 \tab 7 \tab 16 \tab 1 \tab 12 \tab 9 \tab 2 \tab 2 \tab 9 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 4 \tab 3 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 4510 \cr
   2008 \tab 354 \tab 5 \tab 14 \tab 7 \tab - \tab 95 \tab 63 \tab 6 \tab 8 \tab 19 \tab 3 \tab 4 \tab 1 \tab 4 \tab 28 \tab 4 \tab 6 \tab 11 \tab 4 \tab 3 \tab 23 \tab 11 \tab 4 \tab 6 \tab 1 \tab 2 \tab 14 \tab 12 \tab 1 \tab 191 \tab 5 \tab 2 \tab 7 \tab 3 \tab 1 \tab 8 \tab 7 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 26 \tab - \tab 4 \tab 10 \tab 1 \tab 7 \tab 2 \tab 4 \tab 39 \tab 13 \tab 1 \tab 4 \tab 10 \tab - \tab 4 \tab 5 \tab 4 \tab 5 \tab 4 \tab 3 \tab 3 \tab 7 \tab - \tab 6 \tab 8 \tab 10 \tab 31 \tab 10 \tab 2 \tab 10 \tab 23 \tab 3 \tab - \tab 4 \tab - \tab 9 \tab 13 \tab 7 \tab 2 \tab 2 \tab 26 \tab 8 \tab - \tab - \tab 32 \tab 5 \tab - \tab 9 \tab 2 \tab - \tab 8 \tab 4 \tab 4 \tab 7 \tab 21 \tab 7 \tab 6 \tab - \tab 7 \tab 7 \tab 2 \tab 3 \tab 1 \tab 7 \tab - \tab 7 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 3 \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 20 \tab - \tab 37 \tab - \tab 68 \tab 5 \tab 64 \tab 1 \tab 3 \tab 3 \tab 3 \tab 8 \tab 3 \tab 3 \tab 1 \tab 2 \tab 3 \tab 6 \tab 4 \tab 3 \tab 2 \tab 2 \tab 1 \tab 3 \tab - \tab 17 \tab 5 \tab 3 \tab 4 \tab 3 \tab - \tab 2 \tab 8 \tab 2 \tab 10 \tab 1 \tab 3 \tab 3 \tab 31 \tab 19 \tab 12 \tab 4 \tab - \tab 6 \tab 19 \tab 1 \tab 1 \tab 12 \tab 2 \tab 5 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 5 \tab 20 \tab 4 \tab 2 \tab 36 \tab 16 \tab - \tab - \tab 7 \tab 1 \tab 7 \tab 8 \tab 2 \tab 6 \tab 4 \tab 1 \tab - \tab 5 \tab 2 \tab 9 \tab - \tab 2 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 354 \tab 2 \tab 13 \tab 4 \tab 2 \tab 73 \tab 63 \tab 25 \tab 14 \tab 12 \tab 12 \tab 2 \tab 2 \tab 5 \tab 29 \tab 1 \tab 7 \tab 13 \tab 4 \tab 5 \tab 18 \tab 9 \tab - \tab 2 \tab 2 \tab 9 \tab 9 \tab 33 \tab 2 \tab 169 \tab 1 \tab 6 \tab 10 \tab 2 \tab - \tab 5 \tab 8 \tab 1 \tab - \tab 4 \tab 3 \tab 1 \tab 28 \tab - \tab 4 \tab 19 \tab 2 \tab 5 \tab 7 \tab 8 \tab 35 \tab 5 \tab - \tab - \tab 5 \tab 1 \tab 2 \tab 9 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 11 \tab 1 \tab 7 \tab 2 \tab 9 \tab 15 \tab 6 \tab 4 \tab 7 \tab 13 \tab 5 \tab 6 \tab 2 \tab 2 \tab 10 \tab 22 \tab 7 \tab 1 \tab 6 \tab 16 \tab 6 \tab 1 \tab - \tab 26 \tab 6 \tab 1 \tab 6 \tab 2 \tab - \tab 9 \tab 7 \tab 6 \tab 5 \tab 14 \tab 10 \tab 3 \tab 1 \tab 4 \tab 2 \tab 9 \tab 4 \tab 2 \tab 4 \tab 1 \tab 13 \tab 3 \tab 2 \tab 3 \tab - \tab 3 \tab 3 \tab 5 \tab 6 \tab 1 \tab 1 \tab - \tab 4 \tab 2 \tab 30 \tab - \tab 38 \tab 1 \tab 55 \tab 10 \tab 72 \tab 6 \tab 4 \tab 6 \tab 3 \tab 14 \tab 2 \tab 9 \tab 1 \tab - \tab 5 \tab 10 \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab 5 \tab 2 \tab 8 \tab 7 \tab 6 \tab 5 \tab 3 \tab 1 \tab 6 \tab 14 \tab 3 \tab 11 \tab - \tab 2 \tab 1 \tab 34 \tab 20 \tab 8 \tab 3 \tab - \tab 9 \tab 27 \tab 1 \tab - \tab 5 \tab 6 \tab 7 \tab - \tab 8 \tab 1 \tab - \tab 3 \tab 2 \tab 8 \tab 3 \tab 11 \tab 5 \tab 3 \tab 43 \tab 15 \tab - \tab - \tab 5 \tab - \tab 4 \tab 6 \tab 1 \tab 8 \tab 3 \tab 1 \tab - \tab 7 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 9 \tab - \tab 2 \tab 5 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 46072 \tab 71 \tab 96 \tab 70 \tab 32 \tab 1925 \tab 1420 \tab 98 \tab 127 \tab 197 \tab 111 \tab 73 \tab 21 \tab 69 \tab 372 \tab 25 \tab 93 \tab 156 \tab 68 \tab 77 \tab 391 \tab 152 \tab 17 \tab 60 \tab 37 \tab 75 \tab 250 \tab 281 \tab 16 \tab 2846 \tab 79 \tab 26 \tab 139 \tab 28 \tab 8 \tab 114 \tab 112 \tab 26 \tab 37 \tab 57 \tab 51 \tab 19 \tab 329 \tab 11 \tab 40 \tab 254 \tab 31 \tab 149 \tab 61 \tab 82 \tab 587 \tab 141 \tab 25 \tab 37 \tab 89 \tab 30 \tab 35 \tab 62 \tab 55 \tab 54 \tab 26 \tab 31 \tab 22 \tab 160 \tab 15 \tab 93 \tab 61 \tab 149 \tab 388 \tab 76 \tab 52 \tab 147 \tab 330 \tab 80 \tab 53 \tab 47 \tab 32 \tab 101 \tab 227 \tab 81 \tab 56 \tab 64 \tab 305 \tab 182 \tab 26 \tab 4 \tab 451 \tab 93 \tab 13 \tab 134 \tab 19 \tab 7 \tab 131 \tab 96 \tab 59 \tab 109 \tab 238 \tab 140 \tab 57 \tab 33 \tab 59 \tab 110 \tab 83 \tab 60 \tab 22 \tab 51 \tab 17 \tab 104 \tab 45 \tab 35 \tab 13 \tab 10 \tab 53 \tab 22 \tab 110 \tab 61 \tab 35 \tab 12 \tab 12 \tab 19 \tab 34 \tab 278 \tab 27 \tab 589 \tab 11 \tab 847 \tab 107 \tab 867 \tab 32 \tab 48 \tab 28 \tab 32 \tab 173 \tab 21 \tab 51 \tab 29 \tab 33 \tab 56 \tab 115 \tab 38 \tab 55 \tab 57 \tab 32 \tab 16 \tab 47 \tab 6 \tab 208 \tab 74 \tab 154 \tab 39 \tab 49 \tab 17 \tab 74 \tab 172 \tab 29 \tab 155 \tab 30 \tab 43 \tab 28 \tab 599 \tab 204 \tab 216 \tab 38 \tab 7 \tab 110 \tab 233 \tab 24 \tab 12 \tab 65 \tab 39 \tab 63 \tab 9 \tab 116 \tab 18 \tab 20 \tab 20 \tab 40 \tab 55 \tab 25 \tab 208 \tab 55 \tab 53 \tab 594 \tab 300 \tab 8 \tab 13 \tab 61 \tab 8 \tab 71 \tab 90 \tab 9 \tab 171 \tab 50 \tab 14 \tab 8 \tab 69 \tab 24 \tab 62 \tab 13 \tab 17 \tab 25 \tab 13 \tab 8 \tab 65 \tab 20 \tab 10 \tab 5 \tab 57 \tab 24 \tab 26 \tab 38 \tab 12 \tab 23 \tab 19 \tab 19 \tab 1 \tab 5 \tab 11 \tab 11 \tab 8 \tab 4 \tab 9 \tab 14 \tab 5 \tab 6 \tab 24 \tab 7 \tab 2 \tab 12 \tab 4 \tab 1 \tab 2 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{10} agricultural production, crops
\item \code{11} agricultural production, livestock
\item \code{20} agricultural services, except horticultural
\item \code{21} horticultural services
\item \code{30} forestry
\item \code{31} fishing, hunting, and trapping, etc.
\item \code{40} metal mining
\item \code{41} coal mining
\item \code{42} crude petroleum and natural gas extraction
\item \code{50} nonmetallic mining and quarrying, except fuel
\item \code{60} construction
\item \code{100} meat products
\item \code{101} dairy products
\item \code{102} canned and preserved fruits and vegetables
\item \code{110} grain mill products
\item \code{111} bakery products
\item \code{112} sugar and confectionery products
\item \code{120} beverage industries
\item \code{121} miscellaneous food preparations and kindred products
\item \code{122} not specified food industries
\item \code{130} tobacco manufactures
\item \code{132} knitting mills
\item \code{140} dyeing and finishing textiles, except wool and knit goods
\item \code{141} floor coverings, except hard surface
\item \code{142} yarn, thread, and fabric mills
\item \code{150} miscellaneous textile mill products
\item \code{151} apparel and accessories, except knit
\item \code{152} miscellaneous fabricated textile products
\item \code{160} pulp, paper, and paperboard mills
\item \code{161} miscellaneous paper and pulp products
\item \code{162} paperboard containers and boxes
\item \code{171} newspaper publishing and printing
\item \code{172} printing, publishing, and allied industries, except newspapers
\item \code{180} plastics, synthetics, and resins
\item \code{181} drugs
\item \code{182} soaps and cosmetics
\item \code{185} paints, varnishes, and related products
\item \code{190} agricultural chemicals
\item \code{191} industrial and miscellaneous chemicals
\item \code{192} petroleum and coal products
\item \code{200} petroleum refining
\item \code{201} miscellaneous petroleum and coal products
\item \code{210} tires and inner tubes
\item \code{211} other rubber products, and plastics footwear and belting
\item \code{212} miscellaneous plastics products
\item \code{220} leather tanning and finishing
\item \code{221} footwear, except rubber and plastic
\item \code{222} leather products, except footwear
\item \code{230} logging
\item \code{231} sawmills, planning mills, and millwork
\item \code{232} wood buildings and mobile homes
\item \code{241} miscellaneous wood products
\item \code{242} furniture and fixtures
\item \code{250} glass and glass products
\item \code{251} cement, concrete, gypsum, and plaster products
\item \code{252} structural clay products
\item \code{261} pottery and related products
\item \code{262} miscellaneous nonmetallic mineral and stone products
\item \code{270} blast furnaces, steelworks, rolling and finishing mills
\item \code{271} iron and steel foundries
\item \code{272} primary aluminum industries
\item \code{280} other primary metal industries
\item \code{281} cutlery, hand tools, and other hardware
\item \code{282} fabricated structural metal products
\item \code{290} screw machine products
\item \code{291} metal forgings and stampings
\item \code{292} ordnance
\item \code{300} miscellaneous fabricated metal products
\item \code{301} not specified metal industries
\item \code{310} engines and turbines
\item \code{311} farm machinery and equipment
\item \code{312} construction and material handling machines
\item \code{320} metalworking machinery
\item \code{321} office and accounting machines
\item \code{322} electronic computing equipment
\item \code{331} machinery, except electrical, n.e.c.
\item \code{332} not specified machinery
\item \code{340} household appliances
\item \code{341} radio, tv, and communication equipment
\item \code{342} electrical machinery, equipment, and supplies, n.e.c.
\item \code{350} not specified electrical machinery, equipment, and supplies
\item \code{351} motor vehicles and motor vehicle equipment
\item \code{352} aircraft and parts
\item \code{360} ship and boat building and repairing
\item \code{361} railroad locomotives and equipment
\item \code{362} guided missiles, space vehicles, and parts
\item \code{370} cycles and miscellaneous transportation equipment
\item \code{371} scientific and controlling instruments
\item \code{372} optical and health services supplies
\item \code{380} photographic equipment and supplies
\item \code{381} watches, clocks, and clockwork operated devices
\item \code{382} not specified professional equipment
\item \code{390} toys, amusement, and sporting goods
\item \code{391} miscellaneous manufacturing industries
\item \code{392} not specified manufacturing industries
\item \code{400} railroads
\item \code{401} bus service and urban transit
\item \code{402} taxicab service
\item \code{410} trucking service
\item \code{411} warehousing and storage
\item \code{412} u.s. postal service
\item \code{420} water transportation
\item \code{421} air transportation
\item \code{422} pipe lines, except natural gas
\item \code{432} services incidental to transportation
\item \code{440} radio and television broadcasting
\item \code{441} telephone (wire and radio)
\item \code{442} telegraph and miscellaneous communication service
\item \code{460} electric light and power
\item \code{461} gas and steam supply systems
\item \code{462} electric and gas, and other combinations
\item \code{470} water supply and irrigation
\item \code{471} sanitary services
\item \code{472} not specified utilities
\item \code{500} motor vehicles and equipment
\item \code{501} furniture and home furnishings
\item \code{502} lumber and construction materials
\item \code{510} sporting goods, toys, and hobby goods
\item \code{511} metals and minerals, except petroleum
\item \code{512} electrical goods
\item \code{521} hardware, plumbing and heating supplies
\item \code{522} not specified electrical and hardware products
\item \code{530} machinery, equipment, and supplies
\item \code{531} scrap and waste materials
\item \code{532} miscellaneous wholesale, durable goods
\item \code{540} paper and paper products
\item \code{541} drugs, chemicals, and allied products
\item \code{542} apparel, fabrics, and notions
\item \code{550} groceries and related products
\item \code{551} farm products-raw materials
\item \code{552} petroleum products
\item \code{560} alcoholic beverages
\item \code{561} farm supplies
\item \code{562} miscellaneous wholesale, nondurable goods
\item \code{571} not specified wholesale trade
\item \code{580} lumber and building material retailing
\item \code{581} hardware stores
\item \code{582} retail nurseries and garden stores
\item \code{590} mobile home dealers
\item \code{591} department stores
\item \code{592} variety stores
\item \code{600} miscellaneous general merchandise stores
\item \code{601} grocery stores
\item \code{602} dairy products stores
\item \code{610} retail bakeries
\item \code{611} food stores, n.e.c.
\item \code{612} motor vehicle dealers
\item \code{620} auto and home supply stores
\item \code{621} gasoline service stations
\item \code{622} miscellaneous vehicle dealers
\item \code{630} apparel and accessory stores, except shoe
\item \code{631} shoe stores
\item \code{632} furniture and home furnishings stores
\item \code{640} household appliances, tv, and radio stores
\item \code{641} eating and drinking places
\item \code{642} drug stores
\item \code{650} liquor stores
\item \code{651} sporting goods, bicycles, and hobby stores
\item \code{652} book and stationery stores
\item \code{660} jewelry stores
\item \code{661} sewing, needlework, and piece goods stores
\item \code{662} mail order houses
\item \code{670} vending machine operators
\item \code{671} direct selling establishments
\item \code{672} fuel and ice dealers
\item \code{681} retail florists
\item \code{682} miscellaneous retail stores
\item \code{691} not specified retail trade
\item \code{700} banking
\item \code{701} savings and loan associations
\item \code{702} credit agencies, n.e.c.
\item \code{710} security, commodity brokerage, and investment companies
\item \code{711} insurance
\item \code{712} real estate, including real estate-insurance-law offices
\item \code{721} advertising
\item \code{722} services to dwellings and other buildings
\item \code{730} commercial research, development, and testing labs
\item \code{731} personnel supply services
\item \code{732} business management and consulting services
\item \code{740} computer and data processing services
\item \code{741} detective and protective services
\item \code{742} business services, n.e.c.
\item \code{750} automotive services, except repair
\item \code{751} automotive repair shops
\item \code{752} electrical repair shops
\item \code{760} miscellaneous repair services
\item \code{761} private households
\item \code{762} hotels and motels
\item \code{770} lodging places, except hotels and motels
\item \code{771} laundry, cleaning, and garment services
\item \code{772} beauty shops
\item \code{780} barber shops
\item \code{781} funeral service and crematories
\item \code{782} shoe repair shops
\item \code{790} dressmaking shops
\item \code{791} miscellaneous personal services
\item \code{800} theaters and motion pictures
\item \code{801} bowling alleys, billiard and pool parlors
\item \code{802} miscellaneous entertainment and recreation services
\item \code{812} offices of physicians
\item \code{820} offices of dentists
\item \code{821} offices of chiropractors
\item \code{822} offices of optometrists
\item \code{830} offices of health practitioners, n.e.c.
\item \code{831} hospitals
\item \code{832} nursing and personal care facilities
\item \code{840} health services, n.e.c.
\item \code{841} legal services
\item \code{842} elementary and secondary schools
\item \code{850} colleges and universities
\item \code{851} business, trade, and vocational schools
\item \code{852} libraries
\item \code{860} educational services, n.e.c.
\item \code{861} job training and vocational rehabilitation services
\item \code{862} child day care services
\item \code{870} residential care facilities, without nursing
\item \code{871} social services, n.e.c.
\item \code{872} museums, art galleries and zoos
\item \code{880} religious organizations
\item \code{881} membership organizations
\item \code{882} engineering, architectural, and surveying services
\item \code{890} accounting, auditing, and bookkeeping services
\item \code{891} noncommercial educational and scientific research
\item \code{892} miscellaneous professional and related services
\item \code{900} executive and legislative offices
\item \code{901} general government, n.e.c.
\item \code{910} justice, public order, and safety
\item \code{921} public finance, taxation, and monetary policy
\item \code{922} administration of human resources programs
\item \code{930} administration of environmental quality and household programs
\item \code{931} administration of economic programs
\item \code{932} national security and international affairs
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
