% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{paindstatus}
\alias{paindstatus}
\title{paindstatus}
\source{
General Social Survey https://gss.norc.org
}
\description{
Paind10 coding status
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab found, coded \tab found, iap \tab found, no answer \tab not found, coded \tab not found, iap \tab not found, no answer \tab not found, no code assigned \tab iap \tab Total \cr
   1972 \tab 582 \tab 98 \tab 22 \tab 469 \tab 100 \tab 54 \tab 288 \tab - \tab 1613 \cr
   1973 \tab 1087 \tab 126 \tab 21 \tab 127 \tab 25 \tab 7 \tab 111 \tab - \tab 1504 \cr
   1974 \tab 978 \tab 148 \tab 91 \tab 145 \tab 34 \tab 1 \tab 87 \tab - \tab 1484 \cr
   1975 \tab 869 \tab 96 \tab 11 \tab 331 \tab 57 \tab 7 \tab 119 \tab - \tab 1490 \cr
   1976 \tab 1056 \tab 154 \tab 20 \tab 155 \tab 33 \tab 5 \tab 76 \tab - \tab 1499 \cr
   1977 \tab 703 \tab 97 \tab 17 \tab 423 \tab 96 \tab 8 \tab 186 \tab - \tab 1530 \cr
   1978 \tab 1307 \tab 192 \tab 19 \tab 1 \tab 13 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1087 \tab 198 \tab 8 \tab 96 \tab 32 \tab 1 \tab 46 \tab - \tab 1468 \cr
   1982 \tab 1356 \tab 262 \tab 78 \tab 101 \tab 34 \tab 9 \tab 20 \tab - \tab 1860 \cr
   1983 \tab 1275 \tab 182 \tab 18 \tab 76 \tab 25 \tab - \tab 23 \tab - \tab 1599 \cr
   1984 \tab 1247 \tab 194 \tab 17 \tab 4 \tab 8 \tab 1 \tab 2 \tab - \tab 1473 \cr
   1985 \tab 1129 \tab 160 \tab 7 \tab 148 \tab 34 \tab 1 \tab 55 \tab - \tab 1534 \cr
   1986 \tab 1205 \tab 200 \tab 20 \tab 12 \tab 22 \tab 1 \tab 10 \tab - \tab 1470 \cr
   1987 \tab 1256 \tab 237 \tab 18 \tab 161 \tab 54 \tab 11 \tab 82 \tab - \tab 1819 \cr
   1988 \tab 1247 \tab 197 \tab 16 \tab 11 \tab 10 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1211 \tab 196 \tab 13 \tab 78 \tab 29 \tab 4 \tab 6 \tab - \tab 1537 \cr
   1990 \tab 1134 \tab 207 \tab 17 \tab 4 \tab 10 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1245 \tab 211 \tab 43 \tab 4 \tab 9 \tab 4 \tab 1 \tab - \tab 1517 \cr
   1993 \tab 1323 \tab 182 \tab 40 \tab 38 \tab 15 \tab 4 \tab 4 \tab - \tab 1606 \cr
   1994 \tab 2402 \tab 497 \tab 39 \tab 36 \tab 11 \tab 2 \tab 5 \tab - \tab 2992 \cr
   1996 \tab 2325 \tab 495 \tab 63 \tab 12 \tab 7 \tab - \tab 2 \tab - \tab 2904 \cr
   1998 \tab 2249 \tab 488 \tab 64 \tab 11 \tab 15 \tab 2 \tab 3 \tab - \tab 2832 \cr
   2000 \tab 2239 \tab 518 \tab 60 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2175 \tab 526 \tab 64 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2212 \tab 514 \tab 86 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2406 \tab 2049 \tab 55 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1625 \tab 354 \tab 44 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1652 \tab 354 \tab 38 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 40582 \tab 9132 \tab 1009 \tab 2443 \tab 673 \tab 122 \tab 1126 \tab 17303 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} found, iap
\item \code{2} found, no answer
\item \code{3} found, coded
\item \code{4} not found, iap
\item \code{5} not found, no answer
\item \code{6} not found, coded
\item \code{7} not found, no code assigned
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
