% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_595.R
\name{paisc681}
\alias{paisc681}
\title{R's father's occupation, 1980 census}
\source{
General Social Survey https://gss.norc.org
}
\description{
paisc681
}
\details{
Question 1639. Occupation of respondent's father (International Standard Classification Codes-1968)
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 110 \tab 1100 \tab 120 \tab 1219 \tab 131 \tab 1310 \tab 1320 \tab 1330 \tab 1410 \tab 1590 \tab 1610 \tab 1790 \tab 1921 \tab 1939 \tab 1940 \tab 1999 \tab 2010 \tab 2020 \tab 2030 \tab 210 \tab 2110 \tab 2190 \tab 220 \tab 230 \tab 240 \tab 250 \tab 280 \tab 290 \tab 3000 \tab 3102 \tab 320 \tab 3211 \tab 3313 \tab 3319 \tab 340 \tab 3520 \tab 3590 \tab 3609 \tab 3700 \tab 3709 \tab 390 \tab 3911 \tab 3919 \tab 3920 \tab 3939 \tab 3951 \tab 3990 \tab 3993 \tab 4210 \tab 4220 \tab 4222 \tab 4320 \tab 4400 \tab 4410 \tab 4411 \tab 4419 \tab 4420 \tab 4432 \tab 4519 \tab 4520 \tab 4521 \tab 5200 \tab 530 \tab 5319 \tab 540 \tab 5500 \tab 5510 \tab 5600 \tab 5709 \tab 5810 \tab 5829 \tab 5890 \tab 5891 \tab 5895 \tab 5990 \tab 5998 \tab 5999 \tab 6000 \tab 610 \tab 6110 \tab 6120 \tab 6210 \tab 6240 \tab 6270 \tab 630 \tab 6310 \tab 6319 \tab 6410 \tab 670 \tab 7000 \tab 7110 \tab 7111 \tab 7119 \tab 7130 \tab 7220 \tab 7230 \tab 7240 \tab 7270 \tab 7320 \tab 7410 \tab 7430 \tab 7520 \tab 7540 \tab 7730 \tab 7760 \tab 790 \tab 7919 \tab 7990 \tab 8010 \tab 8020 \tab 8110 \tab 8120 \tab 8190 \tab 830 \tab 8311 \tab 8329 \tab 8330 \tab 8340 \tab 8351 \tab 8420 \tab 8430 \tab 8440 \tab 8490 \tab 8493 \tab 8494 \tab 8499 \tab 8520 \tab 8530 \tab 8540 \tab 8559 \tab 8560 \tab 8719 \tab 8720 \tab 8730 \tab 8739 \tab 8740 \tab 8800 \tab 8930 \tab 900 \tab 9210 \tab 9220 \tab 9310 \tab 9390 \tab 9430 \tab 9490 \tab 9499 \tab 9510 \tab 9520 \tab 9530 \tab 9559 \tab 9594 \tab 9599 \tab 9610 \tab 9690 \tab 9710 \tab 9712 \tab 9714 \tab 9730 \tab 9740 \tab 9790 \tab 9810 \tab 9830 \tab 9851 \tab 9852 \tab 9859 \tab 9891 \tab 9971 \tab 9990 \tab 9996 \tab 130 \tab 1390 \tab 1393 \tab 140 \tab 1593 \tab 1621 \tab 1639 \tab 1710 \tab 1799 \tab 1809 \tab 2192 \tab 260 \tab 270 \tab 3800 \tab 410 \tab 4412 \tab 4431 \tab 5312 \tab 5321 \tab 5329 \tab 5991 \tab 5992 \tab 6001 \tab 6320 \tab 650 \tab 7280 \tab 769 \tab 770 \tab 7790 \tab 7950 \tab 8321 \tab 8421 \tab 8422 \tab 8570 \tab 8590 \tab 8910 \tab 9595 \tab 9719 \tab 1730 \tab 1910 \tab 330 \tab 3399 \tab 3809 \tab 4319 \tab 4523 \tab 5821 \tab 5900 \tab 5920 \tab 620 \tab 710 \tab 759 \tab 849 \tab 8610 \tab 8732 \tab 9250 \tab 9560 \tab 9570 \tab 9979 \tab 1490 \tab 310 \tab 3311 \tab 3420 \tab 3959 \tab 5300 \tab 5400 \tab 5702 \tab 690 \tab 7910 \tab 7960 \tab 810 \tab 8620 \tab 9260 \tab 9890 \tab 1229 \tab 3219 \tab 3949 \tab 4900 \tab 5322 \tab 5520 \tab 5700 \tab 8950 \tab 3220 \tab 3701 \tab 6491 \tab 760 \tab 8350 \tab 1924 \tab 5401 \tab 549 \tab 8390 \tab 9240 \tab 1519 \tab 3801 \tab 421 \tab 5320 \tab 510 \tab 9270 \tab 1350 \tab 8990 \tab 1340 \tab 3391 \tab 350 \tab 3999 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 222 \tab 3 \tab 10 \tab 1 \tab 5 \tab 1 \tab 8 \tab 7 \tab 2 \tab 10 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 10 \tab 1 \tab 2 \tab 5 \tab 2 \tab 134 \tab 5 \tab 8 \tab 7 \tab 6 \tab 3 \tab 7 \tab 3 \tab 7 \tab 5 \tab 1 \tab 1 \tab 6 \tab 4 \tab 1 \tab 5 \tab 3 \tab 1 \tab 7 \tab 1 \tab 5 \tab 5 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 35 \tab 6 \tab 1 \tab 11 \tab 2 \tab 1 \tab 2 \tab 7 \tab 1 \tab 2 \tab 30 \tab 4 \tab 3 \tab 9 \tab 1 \tab 4 \tab 1 \tab 12 \tab 1 \tab 3 \tab 4 \tab 4 \tab 8 \tab 10 \tab 1 \tab 19 \tab 1 \tab 2 \tab 1 \tab 2 \tab 7 \tab 190 \tab 7 \tab 31 \tab 1 \tab 10 \tab 3 \tab 1 \tab 6 \tab 1 \tab 1 \tab 42 \tab 5 \tab 1 \tab 9 \tab 3 \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab 6 \tab 2 \tab 2 \tab 1 \tab 5 \tab 4 \tab 1 \tab 4 \tab 4 \tab 4 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 7 \tab 2 \tab 17 \tab 1 \tab 2 \tab 10 \tab 3 \tab 1 \tab 9 \tab 1 \tab 23 \tab 3 \tab 3 \tab 1 \tab 12 \tab 5 \tab 7 \tab 13 \tab 4 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 5 \tab 9 \tab 1 \tab 1 \tab 3 \tab 3 \tab 5 \tab 1 \tab 4 \tab 2 \tab 57 \tab 1 \tab 1 \tab 27 \tab 4 \tab 1 \tab 3 \tab 1 \tab 8 \tab 8 \tab 2 \tab 5 \tab 7 \tab 42 \tab 9 \tab 4 \tab 1 \tab 31 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 242 \tab 3 \tab 8 \tab - \tab 6 \tab - \tab 12 \tab 9 \tab 4 \tab 12 \tab 2 \tab - \tab 2 \tab 1 \tab 5 \tab 4 \tab 3 \tab 5 \tab - \tab 1 \tab 2 \tab 6 \tab 122 \tab 9 \tab 5 \tab 8 \tab 1 \tab 6 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab 2 \tab 6 \tab 3 \tab 8 \tab 10 \tab 3 \tab 5 \tab - \tab 3 \tab 2 \tab - \tab 2 \tab 2 \tab 29 \tab 3 \tab 3 \tab 7 \tab 2 \tab - \tab 6 \tab 10 \tab 1 \tab 2 \tab 20 \tab 7 \tab 1 \tab 6 \tab 1 \tab 4 \tab - \tab 8 \tab 1 \tab 3 \tab 4 \tab 8 \tab 11 \tab 3 \tab 1 \tab 21 \tab - \tab - \tab - \tab 7 \tab 6 \tab 207 \tab 1 \tab 27 \tab 1 \tab 4 \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 45 \tab 3 \tab - \tab 21 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab 7 \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 5 \tab 3 \tab 35 \tab 3 \tab - \tab 19 \tab 1 \tab 11 \tab 7 \tab 2 \tab 17 \tab 6 \tab 6 \tab - \tab 8 \tab 1 \tab 12 \tab 10 \tab 5 \tab 3 \tab 6 \tab - \tab 5 \tab - \tab - \tab 6 \tab 8 \tab 1 \tab 2 \tab 6 \tab 2 \tab 9 \tab - \tab 1 \tab 1 \tab 51 \tab - \tab - \tab 21 \tab 5 \tab 2 \tab 7 \tab 1 \tab 1 \tab 12 \tab 3 \tab 8 \tab 5 \tab 43 \tab 3 \tab 6 \tab 2 \tab 38 \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 235 \tab 2 \tab 8 \tab - \tab 2 \tab - \tab 4 \tab 7 \tab 3 \tab 8 \tab 3 \tab 3 \tab 1 \tab 2 \tab - \tab 7 \tab 2 \tab 9 \tab 1 \tab - \tab 2 \tab 1 \tab 96 \tab 5 \tab 8 \tab 10 \tab 1 \tab 5 \tab 2 \tab 2 \tab 3 \tab 4 \tab 2 \tab - \tab 5 \tab 1 \tab 2 \tab 5 \tab 7 \tab 4 \tab 9 \tab 2 \tab 4 \tab 2 \tab - \tab 2 \tab - \tab 5 \tab - \tab 39 \tab 6 \tab 1 \tab 17 \tab 1 \tab - \tab 5 \tab 9 \tab 2 \tab 1 \tab 25 \tab 6 \tab - \tab 1 \tab 2 \tab 7 \tab - \tab 5 \tab 2 \tab 3 \tab 2 \tab 3 \tab 5 \tab 3 \tab - \tab 13 \tab 4 \tab - \tab 2 \tab 5 \tab 5 \tab 150 \tab - \tab 27 \tab - \tab 8 \tab 2 \tab - \tab 3 \tab - \tab - \tab 31 \tab 4 \tab - \tab 6 \tab 3 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 9 \tab 3 \tab 1 \tab 1 \tab 1 \tab 5 \tab - \tab 4 \tab - \tab 2 \tab 4 \tab - \tab 10 \tab - \tab 25 \tab 2 \tab 1 \tab 17 \tab 2 \tab 10 \tab 14 \tab 4 \tab 19 \tab 4 \tab 1 \tab 2 \tab 9 \tab 8 \tab 9 \tab 20 \tab 3 \tab 3 \tab 2 \tab - \tab 5 \tab 1 \tab 2 \tab 1 \tab 13 \tab 2 \tab 3 \tab 5 \tab 10 \tab 2 \tab - \tab 2 \tab 1 \tab 59 \tab 1 \tab 1 \tab 12 \tab 5 \tab - \tab 1 \tab 7 \tab - \tab 5 \tab - \tab 2 \tab 3 \tab 37 \tab 1 \tab 7 \tab 2 \tab 33 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 266 \tab 4 \tab 12 \tab - \tab 4 \tab - \tab 1 \tab 9 \tab 2 \tab 8 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 7 \tab 1 \tab 1 \tab 3 \tab 3 \tab 103 \tab 5 \tab 13 \tab 7 \tab 1 \tab 8 \tab 6 \tab 4 \tab 5 \tab 4 \tab - \tab - \tab 4 \tab 4 \tab 2 \tab 4 \tab 2 \tab 5 \tab 13 \tab 1 \tab 1 \tab 3 \tab 1 \tab 5 \tab - \tab 2 \tab 3 \tab 45 \tab 2 \tab 2 \tab 26 \tab - \tab 3 \tab 6 \tab 10 \tab - \tab 1 \tab 24 \tab 5 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab 18 \tab 2 \tab 2 \tab 6 \tab 1 \tab 2 \tab 2 \tab 1 \tab 26 \tab - \tab 2 \tab - \tab 32 \tab 10 \tab 163 \tab 1 \tab 16 \tab 2 \tab 9 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 50 \tab - \tab 1 \tab 19 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 5 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 9 \tab 1 \tab 20 \tab 1 \tab 1 \tab 16 \tab 1 \tab 11 \tab 9 \tab 3 \tab 17 \tab - \tab 3 \tab 2 \tab 14 \tab 2 \tab 12 \tab 14 \tab 2 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab 3 \tab 5 \tab 9 \tab 4 \tab 1 \tab 2 \tab 5 \tab 10 \tab 2 \tab - \tab 1 \tab 56 \tab 2 \tab - \tab 12 \tab 6 \tab - \tab 7 \tab 7 \tab 3 \tab 7 \tab - \tab 6 \tab 4 \tab 46 \tab 3 \tab 2 \tab - \tab 28 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 6 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 239 \tab 5 \tab 12 \tab - \tab 10 \tab 2 \tab 7 \tab 10 \tab 6 \tab 11 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 5 \tab 2 \tab 7 \tab 1 \tab 1 \tab 2 \tab 2 \tab 104 \tab 11 \tab 7 \tab 4 \tab 1 \tab 5 \tab 10 \tab 2 \tab 5 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 11 \tab 8 \tab 4 \tab 8 \tab 1 \tab 3 \tab 3 \tab - \tab 2 \tab 2 \tab 5 \tab - \tab 66 \tab 2 \tab - \tab 29 \tab 2 \tab 3 \tab 4 \tab 13 \tab 3 \tab 1 \tab 14 \tab 8 \tab 3 \tab - \tab 3 \tab 3 \tab - \tab 15 \tab 3 \tab 3 \tab 3 \tab 10 \tab 16 \tab 4 \tab - \tab 16 \tab - \tab - \tab 1 \tab 4 \tab 13 \tab 213 \tab - \tab 16 \tab - \tab 7 \tab 4 \tab 1 \tab 5 \tab 2 \tab 2 \tab 49 \tab 7 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 3 \tab 1 \tab - \tab - \tab 6 \tab 3 \tab - \tab 2 \tab 6 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 16 \tab - \tab 14 \tab 2 \tab - \tab 26 \tab 3 \tab 19 \tab 10 \tab 6 \tab 18 \tab 6 \tab 4 \tab 1 \tab 15 \tab 8 \tab 10 \tab 11 \tab 3 \tab 4 \tab 1 \tab - \tab 10 \tab - \tab - \tab 8 \tab 10 \tab 2 \tab 3 \tab 1 \tab 6 \tab 11 \tab 3 \tab 1 \tab 1 \tab 46 \tab - \tab 1 \tab 21 \tab 7 \tab 1 \tab 3 \tab 6 \tab 2 \tab 19 \tab 3 \tab 1 \tab 4 \tab 41 \tab 3 \tab 2 \tab - \tab 39 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 561 \tab 6 \tab 26 \tab 2 \tab 24 \tab - \tab 12 \tab 12 \tab 12 \tab 20 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 10 \tab 1 \tab - \tab 8 \tab 2 \tab 194 \tab 10 \tab 18 \tab 9 \tab 13 \tab 8 \tab 15 \tab 9 \tab 4 \tab 7 \tab 2 \tab - \tab 3 \tab 2 \tab 5 \tab 15 \tab 7 \tab 5 \tab 14 \tab 6 \tab 4 \tab 4 \tab 5 \tab 6 \tab - \tab 7 \tab - \tab 94 \tab 11 \tab - \tab 43 \tab 4 \tab 4 \tab 7 \tab 19 \tab 2 \tab 2 \tab 26 \tab 20 \tab 9 \tab 3 \tab 3 \tab 11 \tab 2 \tab 30 \tab 3 \tab 4 \tab 9 \tab 15 \tab 18 \tab 15 \tab - \tab 31 \tab 3 \tab - \tab 1 \tab 7 \tab 13 \tab 324 \tab - \tab 18 \tab 2 \tab 10 \tab 6 \tab 1 \tab 9 \tab 5 \tab 6 \tab 86 \tab 3 \tab - \tab 18 \tab 6 \tab 1 \tab - \tab 5 \tab 2 \tab 12 \tab 7 \tab - \tab - \tab 5 \tab 13 \tab 2 \tab 3 \tab 4 \tab 4 \tab 5 \tab 1 \tab 2 \tab 1 \tab 3 \tab 6 \tab 3 \tab 8 \tab 5 \tab 32 \tab 5 \tab 2 \tab 30 \tab 5 \tab 14 \tab 20 \tab 5 \tab 66 \tab 6 \tab 8 \tab 1 \tab 21 \tab 13 \tab 18 \tab 28 \tab 7 \tab 8 \tab 6 \tab 1 \tab 19 \tab 1 \tab 2 \tab 6 \tab 19 \tab 5 \tab 6 \tab 13 \tab 11 \tab 12 \tab 5 \tab 8 \tab - \tab 96 \tab 1 \tab 2 \tab 36 \tab 13 \tab 6 \tab 2 \tab 12 \tab 5 \tab 24 \tab 3 \tab 17 \tab 6 \tab 86 \tab 9 \tab 8 \tab 2 \tab 73 \tab 2 \tab 1 \tab 4 \tab - \tab 5 \tab 4 \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 8 \tab 2 \tab 1 \tab - \tab 3 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 6 \tab - \tab 2 \tab - \tab 7 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 5 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 578 \tab 5 \tab 22 \tab 1 \tab 16 \tab 1 \tab 11 \tab 20 \tab 15 \tab 23 \tab 2 \tab 3 \tab - \tab 1 \tab 6 \tab 6 \tab 9 \tab 30 \tab - \tab 1 \tab 4 \tab 10 \tab 211 \tab 15 \tab 27 \tab 13 \tab 9 \tab 10 \tab 7 \tab 2 \tab 14 \tab 9 \tab 4 \tab 2 \tab 9 \tab 2 \tab 1 \tab 12 \tab 11 \tab 4 \tab 14 \tab 3 \tab 2 \tab 9 \tab 4 \tab 5 \tab 1 \tab 7 \tab - \tab 71 \tab 8 \tab 2 \tab 31 \tab 2 \tab 4 \tab 9 \tab 17 \tab - \tab 2 \tab 52 \tab 5 \tab 2 \tab 7 \tab 3 \tab 6 \tab - \tab 25 \tab 5 \tab 7 \tab 10 \tab 8 \tab 16 \tab 12 \tab 7 \tab - \tab 1 \tab - \tab 2 \tab 23 \tab 16 \tab 243 \tab 3 \tab 34 \tab 1 \tab 7 \tab 6 \tab - \tab 6 \tab 6 \tab 5 \tab 94 \tab 4 \tab 1 \tab 13 \tab 4 \tab - \tab - \tab 5 \tab - \tab 7 \tab 3 \tab 2 \tab 3 \tab 2 \tab 11 \tab 8 \tab - \tab 5 \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab - \tab 9 \tab 7 \tab 45 \tab 1 \tab 1 \tab 35 \tab 6 \tab 15 \tab 20 \tab 2 \tab 40 \tab 5 \tab 4 \tab 5 \tab 34 \tab 7 \tab 16 \tab 28 \tab 6 \tab 4 \tab 4 \tab - \tab 6 \tab 2 \tab 2 \tab 3 \tab 13 \tab 3 \tab 3 \tab 4 \tab 5 \tab 10 \tab 5 \tab 3 \tab 2 \tab 125 \tab 2 \tab - \tab 38 \tab 3 \tab - \tab 4 \tab 4 \tab 10 \tab 28 \tab 4 \tab 8 \tab 4 \tab 90 \tab 9 \tab 5 \tab 1 \tab 76 \tab 2 \tab 2 \tab 6 \tab - \tab 1 \tab 2 \tab 3 \tab 6 \tab 2 \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab - \tab 14 \tab 4 \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 2 \tab - \tab 4 \tab - \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab 3 \tab 6 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 571 \tab 5 \tab 28 \tab - \tab 21 \tab 1 \tab 19 \tab 19 \tab 12 \tab 19 \tab 3 \tab 2 \tab 2 \tab 1 \tab 3 \tab 4 \tab - \tab 18 \tab 1 \tab 1 \tab 4 \tab 7 \tab 196 \tab 18 \tab 21 \tab 15 \tab 4 \tab 5 \tab 6 \tab 4 \tab 5 \tab 4 \tab - \tab 1 \tab 6 \tab 7 \tab - \tab 6 \tab 6 \tab 7 \tab 18 \tab 1 \tab 6 \tab 3 \tab 3 \tab 6 \tab - \tab 1 \tab - \tab 82 \tab 5 \tab 1 \tab 26 \tab 4 \tab 5 \tab 9 \tab 23 \tab 1 \tab 1 \tab 39 \tab 9 \tab - \tab 6 \tab 3 \tab 8 \tab 1 \tab 34 \tab 2 \tab 4 \tab 3 \tab 9 \tab 22 \tab 15 \tab 5 \tab 33 \tab 2 \tab - \tab 1 \tab 7 \tab 16 \tab 249 \tab - \tab 27 \tab - \tab 9 \tab 5 \tab - \tab 6 \tab 4 \tab 6 \tab 84 \tab 5 \tab - \tab 25 \tab - \tab - \tab - \tab 6 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 5 \tab 13 \tab 3 \tab 4 \tab 1 \tab 3 \tab 5 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab - \tab 8 \tab 1 \tab 50 \tab 3 \tab 3 \tab 35 \tab 8 \tab 22 \tab 32 \tab 3 \tab 35 \tab 9 \tab 6 \tab 4 \tab 33 \tab 7 \tab 19 \tab 16 \tab 7 \tab 4 \tab 4 \tab 1 \tab 4 \tab 2 \tab - \tab 9 \tab 14 \tab 2 \tab - \tab 5 \tab 10 \tab 17 \tab 1 \tab 2 \tab 3 \tab 134 \tab - \tab 1 \tab 48 \tab 4 \tab 4 \tab 8 \tab 3 \tab 1 \tab 17 \tab 3 \tab 6 \tab 9 \tab 72 \tab 5 \tab 4 \tab - \tab 67 \tab 4 \tab 1 \tab 4 \tab - \tab 5 \tab 1 \tab 3 \tab 3 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 8 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 3 \tab - \tab 9 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 601 \tab 3 \tab 21 \tab 3 \tab 16 \tab 1 \tab 14 \tab 18 \tab 11 \tab 20 \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab 4 \tab 3 \tab 8 \tab 1 \tab 2 \tab 2 \tab 7 \tab 187 \tab 12 \tab 25 \tab 14 \tab 4 \tab 11 \tab 3 \tab 2 \tab 5 \tab 5 \tab - \tab 1 \tab 5 \tab 7 \tab 3 \tab 4 \tab 12 \tab 7 \tab 7 \tab 1 \tab 10 \tab 5 \tab 3 \tab 9 \tab - \tab 3 \tab 4 \tab 67 \tab 5 \tab 3 \tab 39 \tab 3 \tab 5 \tab 12 \tab 17 \tab - \tab 1 \tab 37 \tab 4 \tab 1 \tab 10 \tab - \tab 11 \tab - \tab 33 \tab - \tab 4 \tab 5 \tab 17 \tab 14 \tab 14 \tab 3 \tab 21 \tab - \tab 1 \tab 1 \tab 1 \tab 13 \tab 274 \tab - \tab 34 \tab - \tab 10 \tab 4 \tab - \tab 8 \tab 6 \tab 4 \tab 78 \tab 2 \tab - \tab 21 \tab 2 \tab - \tab - \tab 2 \tab - \tab 7 \tab 6 \tab 3 \tab - \tab 1 \tab 8 \tab 6 \tab - \tab 3 \tab 2 \tab 2 \tab 3 \tab 2 \tab 1 \tab 2 \tab 6 \tab - \tab 7 \tab 1 \tab 62 \tab 2 \tab 1 \tab 40 \tab 7 \tab 8 \tab 21 \tab 4 \tab 38 \tab 7 \tab 10 \tab 2 \tab 26 \tab 5 \tab 22 \tab 21 \tab 5 \tab 7 \tab 12 \tab 3 \tab 4 \tab 1 \tab 1 \tab 12 \tab 13 \tab 2 \tab - \tab 3 \tab 11 \tab 16 \tab - \tab 3 \tab 5 \tab 119 \tab - \tab - \tab 39 \tab 7 \tab 3 \tab 1 \tab 4 \tab 3 \tab 25 \tab 2 \tab 4 \tab 4 \tab 118 \tab 6 \tab 3 \tab - \tab 57 \tab 2 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 5 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab 4 \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 7 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab 6 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 6 \tab - \tab 2 \tab - \tab 2 \tab - \tab 10 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 600 \tab 5 \tab 25 \tab 1 \tab 12 \tab - \tab 12 \tab 19 \tab 10 \tab 16 \tab 3 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2 \tab 5 \tab 13 \tab - \tab 1 \tab 2 \tab 9 \tab 205 \tab 18 \tab 12 \tab 12 \tab 6 \tab 10 \tab 4 \tab 4 \tab 10 \tab 3 \tab 1 \tab - \tab 4 \tab 2 \tab 1 \tab 7 \tab 4 \tab 8 \tab 17 \tab 3 \tab 2 \tab 3 \tab 1 \tab 6 \tab - \tab 5 \tab 4 \tab 94 \tab 4 \tab 1 \tab 40 \tab 2 \tab 5 \tab 8 \tab 13 \tab 3 \tab 2 \tab 18 \tab 7 \tab - \tab 2 \tab 4 \tab 14 \tab 1 \tab 24 \tab 2 \tab 2 \tab 5 \tab 10 \tab 28 \tab 15 \tab 2 \tab 53 \tab 3 \tab 2 \tab 2 \tab 3 \tab 13 \tab 200 \tab - \tab 56 \tab - \tab 11 \tab 4 \tab - \tab 8 \tab 2 \tab 4 \tab 100 \tab 4 \tab 2 \tab 16 \tab 2 \tab - \tab - \tab 4 \tab - \tab 5 \tab 2 \tab 6 \tab 1 \tab 1 \tab 9 \tab 4 \tab 1 \tab - \tab 7 \tab 3 \tab 1 \tab 3 \tab 1 \tab 2 \tab 8 \tab - \tab 17 \tab 2 \tab 76 \tab - \tab 2 \tab 38 \tab 3 \tab 18 \tab 20 \tab 1 \tab 39 \tab 5 \tab 5 \tab 2 \tab 30 \tab 9 \tab 16 \tab 19 \tab 12 \tab 2 \tab 4 \tab 1 \tab 11 \tab 2 \tab - \tab 10 \tab 11 \tab 2 \tab 3 \tab - \tab 11 \tab 14 \tab - \tab - \tab 2 \tab 114 \tab - \tab - \tab 27 \tab 8 \tab 5 \tab 3 \tab 5 \tab 2 \tab 16 \tab 4 \tab 6 \tab 5 \tab 91 \tab 9 \tab 5 \tab - \tab 35 \tab 2 \tab - \tab 3 \tab - \tab 6 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab 5 \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 6 \tab - \tab 2 \tab - \tab 8 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab 2 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 611 \tab 6 \tab 26 \tab 2 \tab 18 \tab - \tab 22 \tab 21 \tab 20 \tab 14 \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 5 \tab 5 \tab 16 \tab - \tab 1 \tab 4 \tab 7 \tab 186 \tab 13 \tab 17 \tab 17 \tab 6 \tab 8 \tab 9 \tab - \tab 9 \tab 4 \tab - \tab - \tab 3 \tab 6 \tab 2 \tab 4 \tab 5 \tab 9 \tab 15 \tab 2 \tab 6 \tab 5 \tab 4 \tab 6 \tab 1 \tab 4 \tab 1 \tab 81 \tab 2 \tab 1 \tab 37 \tab 1 \tab 6 \tab 8 \tab 18 \tab - \tab - \tab 34 \tab 8 \tab 3 \tab 7 \tab 4 \tab 5 \tab 1 \tab 31 \tab 5 \tab 3 \tab 3 \tab 11 \tab 17 \tab 14 \tab 1 \tab 42 \tab 2 \tab 1 \tab 2 \tab 2 \tab 20 \tab 173 \tab 1 \tab 58 \tab - \tab 10 \tab 4 \tab 1 \tab 16 \tab 8 \tab - \tab 94 \tab 5 \tab 1 \tab 23 \tab 5 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 3 \tab 6 \tab 2 \tab 1 \tab 16 \tab 4 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab 14 \tab 1 \tab 13 \tab 3 \tab 61 \tab - \tab - \tab 33 \tab 9 \tab 12 \tab 29 \tab 5 \tab 35 \tab 4 \tab 5 \tab 2 \tab 24 \tab 11 \tab 24 \tab 16 \tab 8 \tab 3 \tab 3 \tab - \tab 8 \tab 2 \tab 3 \tab 7 \tab 15 \tab 2 \tab 5 \tab 2 \tab 8 \tab 13 \tab 2 \tab 4 \tab - \tab 118 \tab 1 \tab - \tab 41 \tab 3 \tab 2 \tab - \tab 4 \tab 3 \tab 17 \tab 3 \tab 2 \tab 11 \tab 81 \tab 11 \tab 5 \tab - \tab 53 \tab 4 \tab 1 \tab 5 \tab - \tab 3 \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 7 \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 2 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab 6 \tab 1 \tab 7 \tab - \tab 6 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2141 \tab 8 \tab 30 \tab 2 \tab 19 \tab 1 \tab 11 \tab 17 \tab 18 \tab 22 \tab 3 \tab - \tab 1 \tab 2 \tab 6 \tab 3 \tab 4 \tab 15 \tab 1 \tab 2 \tab 3 \tab 3 \tab 182 \tab 15 \tab 18 \tab 16 \tab 7 \tab 9 \tab 15 \tab 7 \tab 12 \tab 8 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 12 \tab 11 \tab 10 \tab 15 \tab 5 \tab 4 \tab 2 \tab 1 \tab 3 \tab - \tab 4 \tab 2 \tab 63 \tab 8 \tab 3 \tab 28 \tab 5 \tab 6 \tab 6 \tab 27 \tab - \tab 2 \tab 50 \tab 6 \tab 2 \tab 2 \tab 2 \tab 9 \tab 5 \tab 30 \tab 1 \tab 5 \tab 4 \tab 12 \tab 14 \tab 15 \tab 4 \tab 58 \tab 2 \tab - \tab 1 \tab 6 \tab 20 \tab 200 \tab 2 \tab 56 \tab 2 \tab 10 \tab 7 \tab 2 \tab 21 \tab 8 \tab 6 \tab 99 \tab 5 \tab - \tab 13 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 5 \tab 5 \tab 2 \tab 2 \tab 1 \tab 11 \tab 2 \tab 1 \tab 3 \tab 5 \tab 3 \tab 1 \tab 3 \tab 3 \tab - \tab 6 \tab 1 \tab 8 \tab 1 \tab 67 \tab 3 \tab 2 \tab 38 \tab 3 \tab 18 \tab 25 \tab 4 \tab 43 \tab 7 \tab 9 \tab 4 \tab 34 \tab 10 \tab 21 \tab 24 \tab 9 \tab 4 \tab 5 \tab - \tab 12 \tab 2 \tab 3 \tab 6 \tab 18 \tab 4 \tab 2 \tab 8 \tab 4 \tab 17 \tab 2 \tab 4 \tab 3 \tab 134 \tab - \tab - \tab 29 \tab 2 \tab 2 \tab 6 \tab 6 \tab 3 \tab 20 \tab 2 \tab 8 \tab 9 \tab 90 \tab 4 \tab 4 \tab 1 \tab 59 \tab 5 \tab - \tab 3 \tab - \tab 3 \tab - \tab 9 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 4 \tab 4 \tab 4 \tab 1 \tab 5 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 8 \tab 3 \tab 4 \tab - \tab 3 \tab - \tab 2 \tab 2 \tab 4 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 5 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 407 \tab 3 \tab 21 \tab - \tab 9 \tab - \tab 8 \tab 11 \tab 6 \tab 8 \tab 2 \tab 4 \tab 2 \tab 1 \tab 3 \tab 5 \tab 3 \tab 12 \tab 1 \tab 2 \tab 5 \tab 4 \tab 116 \tab 11 \tab 18 \tab 19 \tab 3 \tab 10 \tab 2 \tab 1 \tab 7 \tab 2 \tab - \tab - \tab 4 \tab 1 \tab 6 \tab 6 \tab 4 \tab 3 \tab 7 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab - \tab 4 \tab 1 \tab 37 \tab 6 \tab 1 \tab 16 \tab 6 \tab 4 \tab 11 \tab 21 \tab - \tab - \tab 22 \tab 6 \tab 3 \tab 7 \tab 2 \tab 8 \tab 2 \tab 30 \tab 3 \tab 2 \tab 4 \tab 13 \tab 13 \tab 13 \tab 3 \tab 45 \tab - \tab 1 \tab 1 \tab 1 \tab 9 \tab 138 \tab 1 \tab 23 \tab - \tab 11 \tab 3 \tab 3 \tab 7 \tab 8 \tab 3 \tab 56 \tab 1 \tab 2 \tab 14 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 7 \tab 2 \tab 1 \tab - \tab 1 \tab 10 \tab 6 \tab 1 \tab 4 \tab 3 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 9 \tab - \tab 12 \tab - \tab 68 \tab 1 \tab 1 \tab 30 \tab 1 \tab 12 \tab 7 \tab 10 \tab 23 \tab 4 \tab - \tab 2 \tab 27 \tab 13 \tab 17 \tab 15 \tab 6 \tab 4 \tab 5 \tab - \tab 2 \tab 4 \tab 1 \tab 4 \tab 8 \tab 2 \tab 5 \tab 1 \tab 2 \tab 5 \tab 3 \tab 4 \tab 3 \tab 70 \tab - \tab 1 \tab 35 \tab 9 \tab 3 \tab 3 \tab 2 \tab 3 \tab 14 \tab 2 \tab 3 \tab 3 \tab 55 \tab 4 \tab 2 \tab - \tab 42 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 5 \tab 2 \tab 4 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 1 \tab 6 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 425 \tab 2 \tab 16 \tab 2 \tab 9 \tab - \tab 15 \tab 14 \tab 8 \tab 6 \tab - \tab - \tab - \tab 2 \tab 5 \tab 3 \tab 8 \tab 6 \tab - \tab - \tab 1 \tab 4 \tab 156 \tab 14 \tab 16 \tab 13 \tab 5 \tab 8 \tab 9 \tab 2 \tab 9 \tab 2 \tab 1 \tab - \tab 2 \tab 7 \tab 1 \tab 8 \tab 3 \tab 5 \tab 5 \tab - \tab 5 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 38 \tab 6 \tab - \tab 15 \tab 4 \tab 7 \tab 5 \tab 11 \tab 2 \tab 1 \tab 43 \tab 3 \tab 1 \tab 11 \tab - \tab 8 \tab 2 \tab 27 \tab 1 \tab 3 \tab 4 \tab 4 \tab 14 \tab 7 \tab 3 \tab 41 \tab 3 \tab - \tab 1 \tab 3 \tab 21 \tab 121 \tab 5 \tab 29 \tab 1 \tab 12 \tab 5 \tab - \tab 7 \tab 9 \tab 3 \tab 80 \tab 1 \tab - \tab 11 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 5 \tab 4 \tab 4 \tab - \tab - \tab 6 \tab 4 \tab - \tab 4 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 8 \tab - \tab 5 \tab 1 \tab 50 \tab 2 \tab 1 \tab 38 \tab 2 \tab 12 \tab 10 \tab 3 \tab 26 \tab 4 \tab 2 \tab 2 \tab 27 \tab 4 \tab 21 \tab 12 \tab 7 \tab 2 \tab - \tab 2 \tab 3 \tab 1 \tab 4 \tab 2 \tab 6 \tab 3 \tab 3 \tab 3 \tab 9 \tab 8 \tab 2 \tab 2 \tab - \tab 74 \tab - \tab - \tab 30 \tab 8 \tab 2 \tab 3 \tab 6 \tab 4 \tab 17 \tab 3 \tab 2 \tab 6 \tab 71 \tab 11 \tab 1 \tab - \tab 25 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab 5 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 4 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 46877 \tab 60 \tab 265 \tab 14 \tab 171 \tab 7 \tab 156 \tab 193 \tab 129 \tab 197 \tab 29 \tab 22 \tab 12 \tab 17 \tab 43 \tab 57 \tab 48 \tab 166 \tab 9 \tab 15 \tab 47 \tab 67 \tab 2192 \tab 161 \tab 213 \tab 164 \tab 67 \tab 106 \tab 97 \tab 43 \tab 97 \tab 60 \tab 14 \tab 8 \tab 55 \tab 51 \tab 28 \tab 105 \tab 86 \tab 80 \tab 159 \tab 30 \tab 59 \tab 46 \tab 32 \tab 58 \tab 6 \tab 52 \tab 19 \tab 841 \tab 74 \tab 19 \tab 365 \tab 38 \tab 53 \tab 98 \tab 215 \tab 15 \tab 18 \tab 434 \tab 98 \tab 30 \tab 74 \tab 29 \tab 101 \tab 16 \tab 322 \tab 31 \tab 48 \tab 66 \tab 125 \tab 198 \tab 142 \tab 31 \tab 419 \tab 21 \tab 9 \tab 16 \tab 103 \tab 182 \tab 2845 \tab 21 \tab 452 \tab 10 \tab 128 \tab 57 \tab 11 \tab 108 \tab 60 \tab 43 \tab 988 \tab 49 \tab 8 \tab 211 \tab 36 \tab 10 \tab 4 \tab 41 \tab 5 \tab 74 \tab 47 \tab 30 \tab 15 \tab 21 \tab 128 \tab 54 \tab 14 \tab 42 \tab 45 \tab 36 \tab 10 \tab 27 \tab 13 \tab 20 \tab 74 \tab 8 \tab 134 \tab 27 \tab 622 \tab 26 \tab 17 \tab 405 \tab 54 \tab 183 \tab 233 \tab 53 \tab 439 \tab 70 \tab 66 \tab 30 \tab 314 \tab 103 \tab 224 \tab 247 \tab 84 \tab 50 \tab 56 \tab 10 \tab 96 \tab 19 \tab 23 \tab 84 \tab 166 \tab 35 \tab 37 \tab 56 \tab 97 \tab 149 \tab 26 \tab 38 \tab 24 \tab 1253 \tab 8 \tab 7 \tab 416 \tab 84 \tab 31 \tab 51 \tab 68 \tab 48 \tab 229 \tab 34 \tab 78 \tab 80 \tab 963 \tab 87 \tab 58 \tab 9 \tab 656 \tab 32 \tab 11 \tab 41 \tab 1 \tab 33 \tab 17 \tab 47 \tab 24 \tab 25 \tab 3 \tab 4 \tab 19 \tab 14 \tab 8 \tab 2 \tab 65 \tab 31 \tab 14 \tab 8 \tab 28 \tab 26 \tab 10 \tab 5 \tab 22 \tab 6 \tab 21 \tab 8 \tab 1 \tab 10 \tab 11 \tab 23 \tab 5 \tab 14 \tab 11 \tab 41 \tab 18 \tab 13 \tab 4 \tab 44 \tab 7 \tab 4 \tab 11 \tab 10 \tab 4 \tab 3 \tab 9 \tab 21 \tab 9 \tab 23 \tab 6 \tab 23 \tab 10 \tab 62 \tab 15 \tab 30 \tab 10 \tab 7 \tab 10 \tab 4 \tab 5 \tab 17 \tab 17 \tab 13 \tab 3 \tab 11 \tab 13 \tab 15 \tab 2 \tab 3 \tab 20 \tab 6 \tab 5 \tab 7 \tab 7 \tab 10 \tab 5 \tab 6 \tab 4 \tab 7 \tab 9 \tab 6 \tab 11 \tab 6 \tab 6 \tab 2 \tab 5 \tab 5 \tab 5 \tab 3 \tab 6 \tab 5 \tab 3 \tab 3 \tab 6 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
