% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_596.R
\name{paisco08}
\alias{paisco08}
\title{R's father's occupation, 2010 census & 2008 isco code}
\source{
General Social Survey https://gss.norc.org
}
\description{
paisco08
}
\details{
Question None
}
\section{Values}{

\itemize{
\item \code{110} commissioned armed forces officers
\item \code{210} non-commissioned armed forces officers
\item \code{310} armed forces occupations, other ranks
\item \code{1000} managers
\item \code{1100} chief executives, senior officials and legislators
\item \code{1110} legislators and senior officials
\item \code{1111} legislators
\item \code{1112} senior government officials
\item \code{1113} traditional chiefs and heads of village
\item \code{1114} senior officials of special-interest organizations
\item \code{1120} managing directors and chief executives
\item \code{1200} administrative and commercial managers
\item \code{1210} business services and administration managers
\item \code{1211} finance managers
\item \code{1212} human resource managers
\item \code{1213} policy and planning managers
\item \code{1219} business services and administration managers not elsewhere classified
\item \code{1220} sales, marketing and development managers
\item \code{1221} sales and marketing managers
\item \code{1222} advertising and public relations managers
\item \code{1223} research and development managers
\item \code{1300} production and specialised services managers
\item \code{1310} production managers in agriculture, forestry and fisheries
\item \code{1311} agricultural and forestry production managers
\item \code{1312} aquaculture and fisheries production managers
\item \code{1320} manufacturing, mining, construction, and distribution managers
\item \code{1321} manufacturing managers
\item \code{1322} mining managers
\item \code{1323} construction managers
\item \code{1324} supply, distribution and related managers
\item \code{1330} information and communications technology service managers
\item \code{1340} professional services managers
\item \code{1341} child care services managers
\item \code{1342} health services managers
\item \code{1343} aged care services managers
\item \code{1344} social welfare managers
\item \code{1345} education managers
\item \code{1346} financial and insurance services branch managers
\item \code{1349} professional services managers not elsewhere classified
\item \code{1400} hospitality, retail and other services managers
\item \code{1410} hotel and restaurant managers
\item \code{1411} hotel managers
\item \code{1412} restaurant managers
\item \code{1420} retail and wholesale trade managers
\item \code{1430} other services managers
\item \code{1431} sports, recreation and cultural centre managers
\item \code{1439} services managers not elsewhere classified
\item \code{2000} professionals
\item \code{2100} science and engineering professionals
\item \code{2110} physical and earth science professionals
\item \code{2111} physicists and astronomers
\item \code{2112} meteorologists
\item \code{2113} chemists
\item \code{2114} geologists and geophysicists
\item \code{2120} mathematicians, actuaries and statisticians
\item \code{2130} life science professionals
\item \code{2131} biologists, botanists, zoologists and related professionals
\item \code{2132} farming, forestry and fisheries advisers
\item \code{2133} environmental protection professionals
\item \code{2140} engineering professionals (excluding electrotechnology)
\item \code{2141} industrial and production engineers
\item \code{2142} civil engineers
\item \code{2143} environmental engineers
\item \code{2144} mechanical engineers
\item \code{2145} chemical engineers
\item \code{2146} mining engineers, metallurgists and related professionals
\item \code{2149} engineering professionals not elsewhere classified
\item \code{2150} electrotechnology engineers
\item \code{2151} electrical engineers
\item \code{2152} electronics engineers
\item \code{2153} telecommunications engineers
\item \code{2160} architects, planners, surveyors and designers
\item \code{2161} building architects
\item \code{2162} landscape architects
\item \code{2163} product and garment designers
\item \code{2164} town and traffic planners
\item \code{2165} cartographers and surveyors
\item \code{2166} graphic and multimedia designers
\item \code{2200} health professionals
\item \code{2210} medical doctors
\item \code{2211} generalist medical practitioners
\item \code{2212} specialist medical practitioners
\item \code{2220} nursing and midwifery professionals
\item \code{2221} nursing professionals
\item \code{2222} midwifery professionals
\item \code{2230} traditional and complementary medicine professionals
\item \code{2240} paramedical practitioners
\item \code{2250} veterinarians
\item \code{2260} other health professionals
\item \code{2261} dentists
\item \code{2262} pharmacists
\item \code{2263} environmental and occupational health and hygiene professionals
\item \code{2264} physiotherapists
\item \code{2265} dieticians and nutritionists
\item \code{2266} audiologists and speech therapists
\item \code{2267} optometrists and ophthalmic opticians
\item \code{2269} health professionals not elsewhere classified
\item \code{2300} teaching professionals
\item \code{2310} university and higher education teachers
\item \code{2320} vocational education teachers
\item \code{2330} secondary education teachers
\item \code{2340} primary school and early childhood teachers
\item \code{2341} primary school teachers
\item \code{2342} early childhood educators
\item \code{2350} other teaching professionals
\item \code{2351} education methods specialists
\item \code{2352} special needs teachers
\item \code{2353} other language teachers
\item \code{2354} other music teachers
\item \code{2355} other arts teachers
\item \code{2356} information technology trainers
\item \code{2359} teaching professionals not elsewhere classified
\item \code{2400} business and administration professionals
\item \code{2410} finance professionals
\item \code{2411} accountants
\item \code{2412} financial and investment advisers
\item \code{2413} financial analysts
\item \code{2420} administration professionals
\item \code{2421} management and organization analysts
\item \code{2422} policy administration professionals
\item \code{2423} personnel and careers professionals
\item \code{2424} training and staff development professionals
\item \code{2430} sales, marketing and public relations professionals
\item \code{2431} advertising and marketing professionals
\item \code{2432} public relations professionals
\item \code{2433} technical and medical sales professionals (excluding ict)
\item \code{2434} information and communications technology sales professionals
\item \code{2500} information and communications technology professionals
\item \code{2510} software and applications developers and analysts
\item \code{2511} systems analysts
\item \code{2512} software developers
\item \code{2513} web and multimedia developers
\item \code{2514} applications programmers
\item \code{2519} software and applications developers and analysts not elsewhere classified
\item \code{2520} database and network professionals
\item \code{2521} database designers and administrators
\item \code{2522} systems administrators
\item \code{2523} computer network professionals
\item \code{2529} database and network professionals not elsewhere classified
\item \code{2600} legal, social and cultural professionals
\item \code{2610} legal professionals
\item \code{2611} lawyers
\item \code{2612} judges
\item \code{2619} legal professionals not elsewhere classified
\item \code{2620} librarians, archivists and curators
\item \code{2621} archivists and curators
\item \code{2622} librarians and related information professionals
\item \code{2630} social and religious professionals
\item \code{2631} economists
\item \code{2632} sociologists, anthropologists and related professionals
\item \code{2633} philosophers, historians and political scientists
\item \code{2634} psychologists
\item \code{2635} social work and counselling professionals
\item \code{2636} religious professionals
\item \code{2640} authors, journalists and linguists
\item \code{2641} authors and related writers
\item \code{2642} journalists
\item \code{2643} translators, interpreters and other linguists
\item \code{2650} creative and performing artists
\item \code{2651} visual artists
\item \code{2652} musicians, singers and composers
\item \code{2653} dancers and choreographers
\item \code{2654} film, stage and related directors and producers
\item \code{2655} actors
\item \code{2656} announcers on radio, television and other media
\item \code{2659} creative and performing artists not elsewhere classified
\item \code{3000} technicians and associate professionals
\item \code{3100} science and engineering associate professionals
\item \code{3110} physical and engineering science technicians
\item \code{3111} chemical and physical science technicians
\item \code{3112} civil engineering technicians
\item \code{3113} electrical engineering technicians
\item \code{3114} electronics engineering technicians
\item \code{3115} mechanical engineering technicians
\item \code{3116} chemical engineering technicians
\item \code{3117} mining and metallurgical technicians
\item \code{3118} draughtspersons
\item \code{3119} physical and engineering science technicians not elsewhere classified
\item \code{3120} mining, manufacturing and construction supervisors
\item \code{3121} mining supervisors
\item \code{3122} manufacturing supervisors
\item \code{3123} construction supervisors
\item \code{3130} process control technicians
\item \code{3131} power production plant operators
\item \code{3132} incinerator and water treatment plant operators
\item \code{3133} chemical processing plant controllers
\item \code{3134} petroleum and natural gas refining plant operators
\item \code{3135} metal production process controllers
\item \code{3139} process control technicians not elsewhere classified
\item \code{3140} life science technicians and related associate professionals
\item \code{3141} life science technicians (excluding medical)
\item \code{3142} agricultural technicians
\item \code{3143} forestry technicians
\item \code{3150} ship and aircraft controllers and technicians
\item \code{3151} ships' engineers
\item \code{3152} ships' deck officers and pilots
\item \code{3153} aircraft pilots and related associate professionals
\item \code{3154} air traffic controllers
\item \code{3155} air traffic safety electronics technicians
\item \code{3200} health associate professionals
\item \code{3210} medical and pharmaceutical technicians
\item \code{3211} medical imaging and therapeutic equipment technicians
\item \code{3212} medical and pathology laboratory technicians
\item \code{3213} pharmaceutical technicians and assistants
\item \code{3214} medical and dental prosthetic technicians
\item \code{3220} nursing and midwifery associate professionals
\item \code{3221} nursing associate professionals
\item \code{3222} midwifery associate professionals
\item \code{3230} traditional and complementary medicine associate professionals
\item \code{3240} veterinary technicians and assistants
\item \code{3250} other health associate professionals
\item \code{3251} dental assistants and therapists
\item \code{3252} medical records and health information technicians
\item \code{3253} community health workers
\item \code{3254} dispensing opticians
\item \code{3255} physiotherapy technicians and assistants
\item \code{3256} medical assistants
\item \code{3257} environmental and occupational health inspectors and associates
\item \code{3258} ambulance workers
\item \code{3259} health associate professionals not elsewhere classified
\item \code{3300} business and administration associate professionals
\item \code{3310} financial and mathematical associate professionals
\item \code{3311} securities and finance dealers and brokers
\item \code{3312} credit and loans officers
\item \code{3313} accounting associate professionals
\item \code{3314} statistical, mathematical and related associate professionals
\item \code{3315} valuers and loss assessors
\item \code{3320} sales and purchasing agents and brokers
\item \code{3321} insurance representatives
\item \code{3322} commercial sales representatives
\item \code{3323} buyers
\item \code{3324} trade brokers
\item \code{3330} business services agents
\item \code{3331} clearing and forwarding agents
\item \code{3332} conference and event planners
\item \code{3333} employment agents and contractors
\item \code{3334} real estate agents and property managers
\item \code{3339} business services agents not elsewhere classified
\item \code{3340} administrative and specialised secretaries
\item \code{3341} office supervisors
\item \code{3342} legal secretaries
\item \code{3343} administrative and executive secretaries
\item \code{3344} medical secretaries
\item \code{3350} regulatory government associate professionals
\item \code{3351} customs and border inspectors
\item \code{3352} government tax and excise officials
\item \code{3353} government social benefits officials
\item \code{3354} government licensing officials
\item \code{3355} police inspectors and detectives
\item \code{3359} regulatory government associate professionals not elsewhere classified
\item \code{3400} legal, social, cultural and related associate professionals
\item \code{3410} legal, social and religious associate professionals
\item \code{3411} police inspectors and detectives
\item \code{3412} social work associate professionals
\item \code{3413} religious associate professionals
\item \code{3420} sports and fitness workers
\item \code{3421} athletes and sports players
\item \code{3422} sports coaches, instructors and officials
\item \code{3423} fitness and recreation instructors and program leaders
\item \code{3430} artistic, cultural and culinary associate professionals
\item \code{3431} photographers
\item \code{3432} interior designers and decorators
\item \code{3433} gallery, museum and library technicians
\item \code{3434} chefs
\item \code{3435} other artistic and cultural associate professionals
\item \code{3500} information and communications technicians
\item \code{3510} information and communications technology operations and user support technicians
\item \code{3511} information and communications technology operations technicians
\item \code{3512} information and communications technology user support technicians
\item \code{3513} computer network and systems technicians
\item \code{3514} web technicians
\item \code{3520} telecommunications and broadcasting technicians
\item \code{3521} broadcasting and audio-visual technicians
\item \code{3522} telecommunications engineering technicians
\item \code{4000} clerical support workers
\item \code{4100} general and keyboard clerks
\item \code{4110} general office clerks
\item \code{4120} secretaries (general)
\item \code{4130} keyboard operators
\item \code{4131} typists and word processing operators
\item \code{4132} data entry clerks
\item \code{4200} customer services clerks
\item \code{4210} tellers, money collectors and related clerks
\item \code{4211} bank tellers and related clerks
\item \code{4212} bookmakers, croupiers and related gaming workers
\item \code{4213} pawnbrokers and money-lenders
\item \code{4214} debt-collectors and related workers
\item \code{4220} client information workers
\item \code{4221} travel consultants and clerks
\item \code{4222} contact centre information clerks
\item \code{4223} telephone switchboard operators
\item \code{4224} hotel receptionists
\item \code{4225} enquiry clerks
\item \code{4226} receptionists (general)
\item \code{4227} survey and market research interviewers
\item \code{4229} client information workers not elsewhere classified
\item \code{4300} numerical and material recording clerks
\item \code{4310} numerical clerks
\item \code{4311} accounting and bookkeeping clerks
\item \code{4312} statistical, finance and insurance clerks
\item \code{4313} payroll clerks
\item \code{4320} material-recording and transport clerks
\item \code{4321} stock clerks
\item \code{4322} production clerks
\item \code{4323} transport clerks
\item \code{4400} other clerical support workers
\item \code{4410} other clerical support workers
\item \code{4411} library clerks
\item \code{4412} mail carriers and sorting clerks
\item \code{4413} coding, proof-reading and related clerks
\item \code{4414} scribes and related workers
\item \code{4415} filing and copying clerks
\item \code{4416} personnel clerks
\item \code{4419} clerical support workers not elsewhere classified
\item \code{5000} service and sales workers
\item \code{5100} personal service workers
\item \code{5110} travel attendants, conductors and guides
\item \code{5111} travel attendants and travel stewards
\item \code{5112} transport conductors
\item \code{5113} travel guides
\item \code{5120} cooks
\item \code{5130} waiters and bartenders
\item \code{5131} waiters
\item \code{5132} bartenders
\item \code{5140} hairdressers, beauticians and related workers
\item \code{5141} hairdressers
\item \code{5142} beauticians and related workers
\item \code{5150} building and housekeeping supervisors
\item \code{5151} cleaning and housekeeping supervisors in offices, hotels and other establishments
\item \code{5152} domestic housekeepers
\item \code{5153} building caretakers
\item \code{5160} other personal services workers
\item \code{5161} astrologers, fortune-tellers and related workers
\item \code{5162} companions and valets
\item \code{5163} undertakers and embalmers
\item \code{5164} pet groomers and animal care workers
\item \code{5165} driving instructors
\item \code{5169} personal services workers not elsewhere classified
\item \code{5200} sales workers
\item \code{5210} street and market salespersons
\item \code{5211} stall and market salespersons
\item \code{5212} street food salespersons
\item \code{5220} shop salespersons
\item \code{5221} shop keepers
\item \code{5222} shop supervisors
\item \code{5223} shop sales assistants
\item \code{5230} cashiers and ticket clerks
\item \code{5240} other sales workers
\item \code{5241} fashion and other models
\item \code{5242} sales demonstrators
\item \code{5243} door to door salespersons
\item \code{5244} contact centre salespersons
\item \code{5245} service station attendants
\item \code{5246} food service counter attendants
\item \code{5249} sales workers not elsewhere classified
\item \code{5300} personal care workers
\item \code{5310} child care workers and teachers' aides
\item \code{5311} child care workers
\item \code{5312} teachers' aides
\item \code{5320} personal care workers in health services
\item \code{5321} health care assistants
\item \code{5322} home-based personal care workers
\item \code{5329} personal care workers in health services not elsewhere classified
\item \code{5400} protective services workers
\item \code{5410} protective services workers
\item \code{5411} fire-fighters
\item \code{5412} police officers
\item \code{5413} prison guards
\item \code{5414} security guards
\item \code{5419} protective services workers not elsewhere classified
\item \code{6000} skilled agricultural, forestry and fishery workers
\item \code{6100} market-oriented skilled agricultural workers
\item \code{6110} market gardeners and crop growers
\item \code{6111} field crop and vegetable growers
\item \code{6112} tree and shrub crop growers
\item \code{6113} gardeners, horticultural and nursery growers
\item \code{6114} mixed crop growers
\item \code{6120} animal producers
\item \code{6121} livestock and dairy producers
\item \code{6122} poultry producers
\item \code{6123} apiarists and sericulturists
\item \code{6129} animal producers not elsewhere classified
\item \code{6130} mixed crop and animal producers
\item \code{6200} market-oriented skilled forestry, fishery and hunting workers
\item \code{6210} forestry and related workers
\item \code{6220} fishery workers, hunters and trappers
\item \code{6221} aquaculture workers
\item \code{6222} inland and coastal waters fishery workers
\item \code{6223} deep-sea fishery workers
\item \code{6224} hunters and trappers
\item \code{6300} subsistence farmers, fishers, hunters and gatherers
\item \code{6310} subsistence crop farmers
\item \code{6320} subsistence livestock farmers
\item \code{6330} subsistence mixed crop and livestock farmers
\item \code{6340} subsistence fishers, hunters, trappers and gatherers
\item \code{7000} craft and related trades workers
\item \code{7100} building and related trades workers, excluding electricians
\item \code{7110} building frame and related trades workers
\item \code{7111} house builders
\item \code{7112} bricklayers and related workers
\item \code{7113} stonemasons, stone cutters, splitters and carvers
\item \code{7114} concrete placers, concrete finishers and related workers
\item \code{7115} carpenters and joiners
\item \code{7119} building frame and related trades workers not elsewhere classified
\item \code{7120} building finishers and related trades workers
\item \code{7121} roofers
\item \code{7122} floor layers and tile setters
\item \code{7123} plasterers
\item \code{7124} insulation workers
\item \code{7125} glaziers
\item \code{7126} plumbers and pipe fitters
\item \code{7127} air conditioning and refrigeration mechanics
\item \code{7130} painters, building structure cleaners and related trades workers
\item \code{7131} painters and related workers
\item \code{7132} spray painters and varnishers
\item \code{7133} building structure cleaners
\item \code{7200} metal, machinery and related trades workers
\item \code{7210} sheet and structural metal workers, moulders and welders, and related workers
\item \code{7211} metal moulders and coremakers
\item \code{7212} welders and flamecutters
\item \code{7213} sheet-metal workers
\item \code{7214} structural-metal preparers and erectors
\item \code{7215} riggers and cable splicers
\item \code{7220} blacksmiths, toolmakers and related trades workers
\item \code{7221} blacksmiths, hammersmiths and forging press workers
\item \code{7222} toolmakers and related workers
\item \code{7223} metal working machine tool setters and operators
\item \code{7224} metal polishers, wheel grinders and tool sharpeners
\item \code{7230} machinery mechanics and repairers
\item \code{7231} motor vehicle mechanics and repairers
\item \code{7232} aircraft engine mechanics and repairers
\item \code{7233} agricultural and industrial machinery mechanics and repairers
\item \code{7234} bicycle and related repairers
\item \code{7300} handicraft and printing workers
\item \code{7310} handicraft workers
\item \code{7311} precision-instrument makers and repairers
\item \code{7312} musical instrument makers and tuners
\item \code{7313} jewellery and precious-metal workers
\item \code{7314} potters and related workers
\item \code{7315} glass makers, cutters, grinders and finishers
\item \code{7316} sign writers, decorative painters, engravers and etchers
\item \code{7317} handicraft workers in wood, basketry and related materials
\item \code{7318} handicraft workers in textile, leather and related materials
\item \code{7319} handicraft workers not elsewhere classified
\item \code{7320} printing trades workers
\item \code{7321} pre-press technicians
\item \code{7322} printers
\item \code{7323} print finishing and binding workers
\item \code{7400} electrical and electronic trades workers
\item \code{7410} electrical equipment installers and repairers
\item \code{7411} building and related electricians
\item \code{7412} electrical mechanics and fitters
\item \code{7413} electrical line installers and repairers
\item \code{7420} electronics and telecommunications installers and repairers
\item \code{7421} electronics mechanics and servicers
\item \code{7422} information and communications technology installers and servicers
\item \code{7500} food processing, wood working, garment and other craft and related trades workers
\item \code{7510} food processing and related trades workers
\item \code{7511} butchers, fishmongers and related food preparers
\item \code{7512} bakers, pastry-cooks and confectionery makers
\item \code{7513} dairy-products makers
\item \code{7514} fruit, vegetable and related preservers
\item \code{7515} food and beverage tasters and graders
\item \code{7516} tobacco preparers and tobacco products makers
\item \code{7520} wood treaters, cabinet-makers and related trades workers
\item \code{7521} wood treaters
\item \code{7522} cabinet-makers and related workers
\item \code{7523} woodworking-machine tool setters and operators
\item \code{7530} garment and related trades workers
\item \code{7531} tailors, dressmakers, furriers and hatters
\item \code{7532} garment and related pattern-makers and cutters
\item \code{7533} sewing, embroidery and related workers
\item \code{7534} upholsterers and related workers
\item \code{7535} pelt dressers, tanners and fellmongers
\item \code{7536} shoemakers and related workers
\item \code{7540} other craft and related workers
\item \code{7541} underwater divers
\item \code{7542} shotfirers and blasters
\item \code{7543} product graders and testers (excluding foods and beverages)
\item \code{7544} fumigators and other pest and weed controllers
\item \code{7549} craft and related workers not elsewhere classified
\item \code{8000} plant and machine operators, and assemblers
\item \code{8100} stationary plant and machine operators
\item \code{8110} mining and mineral processing plant operators
\item \code{8111} miners and quarriers
\item \code{8112} mineral and stone processing plant operators
\item \code{8113} well drillers and borers and related workers
\item \code{8114} cement, stone and other mineral products machine operators
\item \code{8120} metal processing and finishing plant operators
\item \code{8121} metal processing plant operators
\item \code{8122} metal finishing, plating and coating machine operators
\item \code{8130} chemical and photographic products plant and machine operators
\item \code{8131} chemical products plant and machine operators
\item \code{8132} photographic products machine operators
\item \code{8140} rubber, plastic and paper products machine operators
\item \code{8141} rubber products machine operators
\item \code{8142} plastic products machine operators
\item \code{8143} paper products machine operators
\item \code{8150} textile, fur and leather products machine operators
\item \code{8151} fibre preparing, spinning and winding machine operators
\item \code{8152} weaving and knitting machine operators
\item \code{8153} sewing machine operators
\item \code{8154} bleaching, dyeing and fabric cleaning machine operators
\item \code{8155} fur and leather preparing machine operators
\item \code{8156} shoemaking and related machine operators
\item \code{8157} laundry machine operators
\item \code{8159} textile, fur and leather products machine operators not elsewhere classified
\item \code{8160} food and related products machine operators
\item \code{8170} wood processing and papermaking plant operators
\item \code{8171} pulp and papermaking plant operators
\item \code{8172} wood processing plant operators
\item \code{8180} other stationary plant and machine operators
\item \code{8181} glass and ceramics plant operators
\item \code{8182} steam engine and boiler operators
\item \code{8183} packing, bottling and labelling machine operators
\item \code{8189} stationary plant and machine operators not elsewhere classified
\item \code{8200} assemblers
\item \code{8210} assemblers
\item \code{8211} mechanical machinery assemblers
\item \code{8212} electrical and electronic equipment assemblers
\item \code{8219} assemblers not elsewhere classified
\item \code{8300} drivers and mobile plant operators
\item \code{8310} locomotive engine drivers and related workers
\item \code{8311} locomotive engine drivers
\item \code{8312} railway brake, signal and switch operators
\item \code{8320} car, van and motorcycle drivers
\item \code{8321} motorcycle drivers
\item \code{8322} car, taxi and van drivers
\item \code{8330} heavy truck and bus drivers
\item \code{8331} bus and tram drivers
\item \code{8332} heavy truck and lorry drivers
\item \code{8340} mobile plant operators
\item \code{8341} mobile farm and forestry plant operators
\item \code{8342} earthmoving and related plant operators
\item \code{8343} crane, hoist and related plant operators
\item \code{8344} lifting truck operators
\item \code{8350} ships' deck crews and related workers
\item \code{9000} elementary occupations
\item \code{9100} cleaners and helpers
\item \code{9110} domestic, hotel and office cleaners and helpers
\item \code{9111} domestic cleaners and helpers
\item \code{9112} cleaners and helpers in offices, hotels and other establishments
\item \code{9120} vehicle, window, laundry and other hand cleaning workers
\item \code{9121} hand launderers and pressers
\item \code{9122} vehicle cleaners
\item \code{9123} window cleaners
\item \code{9129} other cleaning workers
\item \code{9200} agricultural, forestry and fishery labourers
\item \code{9210} agricultural, forestry and fishery labourers
\item \code{9211} crop farm labourers
\item \code{9212} livestock farm labourers
\item \code{9213} mixed crop and livestock farm labourers
\item \code{9214} garden and horticultural labourers
\item \code{9215} forestry labourers
\item \code{9216} fishery and aquaculture labourers
\item \code{9300} labourers in mining, construction, manufacturing and transport
\item \code{9310} mining and construction labourers
\item \code{9311} mining and quarrying labourers
\item \code{9312} civil engineering labourers
\item \code{9313} building construction labourers
\item \code{9320} manufacturing labourers
\item \code{9321} hand packers
\item \code{9329} manufacturing labourers not elsewhere classified
\item \code{9330} transport and storage labourers
\item \code{9331} hand and pedal vehicle drivers
\item \code{9332} drivers of animal-drawn vehicles and machinery
\item \code{9333} freight handlers
\item \code{9334} shelf fillers
\item \code{9400} food preparation assistants
\item \code{9410} food preparation assistants
\item \code{9411} fast food preparers
\item \code{9412} kitchen helpers
\item \code{9500} street and related sales and service workers
\item \code{9510} street and related service workers
\item \code{9520} street vendors (excluding food)
\item \code{9600} refuse workers and other elementary workers
\item \code{9610} refuse workers
\item \code{9611} garbage and recycling collectors
\item \code{9612} refuse sorters
\item \code{9613} sweepers and related labourers
\item \code{9620} other elementary workers
\item \code{9621} messengers, package deliverers and luggage porters
\item \code{9622} odd job persons
\item \code{9623} meter readers and vending-machine collectors
\item \code{9624} water and firewood collectors
\item \code{9629} elementary workers not elsewhere classified
\item \code{9997} uncodeable
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
