% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_595.R
\name{paisco68}
\alias{paisco68}
\title{paisco68}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's father's occupation, 1970 census
}
\details{
Question 1636. Occupation of respondent's father (International Standard Classification Codes-1968)
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 110 \tab 1100 \tab 1210 \tab 1220 \tab 131 \tab 1310 \tab 1320 \tab 1330 \tab 1391 \tab 1410 \tab 1591 \tab 1593 \tab 1621 \tab 1630 \tab 1710 \tab 1791 \tab 1921 \tab 1990 \tab 210 \tab 2114 \tab 2190 \tab 220 \tab 230 \tab 240 \tab 250 \tab 260 \tab 270 \tab 280 \tab 3000 \tab 3100 \tab 3102 \tab 320 \tab 3210 \tab 3310 \tab 3311 \tab 3313 \tab 3314 \tab 3390 \tab 3420 \tab 3520 \tab 3590 \tab 3600 \tab 3601 \tab 3700 \tab 3701 \tab 3801 \tab 3910 \tab 3911 \tab 3990 \tab 3991 \tab 3993 \tab 410 \tab 420 \tab 4210 \tab 4220 \tab 4221 \tab 4222 \tab 4410 \tab 4411 \tab 4412 \tab 4432 \tab 4510 \tab 4521 \tab 5000 \tab 530 \tab 5321 \tab 5510 \tab 5520 \tab 5600 \tab 5700 \tab 5810 \tab 5820 \tab 5822 \tab 5890 \tab 5892 \tab 5895 \tab 5920 \tab 5992 \tab 6001 \tab 610 \tab 6110 \tab 6210 \tab 630 \tab 6410 \tab 670 \tab 7000 \tab 7110 \tab 720 \tab 7210 \tab 7320 \tab 7321 \tab 7410 \tab 7520 \tab 7530 \tab 7590 \tab 7710 \tab 7731 \tab 7740 \tab 7760 \tab 791 \tab 7910 \tab 7920 \tab 7940 \tab 7950 \tab 7960 \tab 7990 \tab 8010 \tab 8020 \tab 8110 \tab 8310 \tab 8320 \tab 8330 \tab 8340 \tab 8351 \tab 8400 \tab 8410 \tab 8412 \tab 8430 \tab 8490 \tab 8493 \tab 8494 \tab 8540 \tab 8550 \tab 8560 \tab 8570 \tab 8590 \tab 8710 \tab 8720 \tab 8730 \tab 8732 \tab 8740 \tab 8800 \tab 8910 \tab 900 \tab 9211 \tab 9240 \tab 9250 \tab 9310 \tab 9390 \tab 9490 \tab 9510 \tab 9520 \tab 9530 \tab 9540 \tab 9550 \tab 9590 \tab 9595 \tab 9690 \tab 9710 \tab 9711 \tab 9714 \tab 9730 \tab 9740 \tab 9790 \tab 9810 \tab 9830 \tab 9831 \tab 9840 \tab 9850 \tab 9851 \tab 9852 \tab 9853 \tab 9950 \tab 9971 \tab 9990 \tab 9996 \tab iap \tab 140 \tab 1490 \tab 1610 \tab 1790 \tab 1800 \tab 1930 \tab 2197 \tab 290 \tab 310 \tab 3315 \tab 350 \tab 390 \tab 3940 \tab 3992 \tab 4002 \tab 4420 \tab 4431 \tab 5103 \tab 5201 \tab 5400 \tab 5993 \tab 6000 \tab 6117 \tab 6290 \tab 650 \tab 710 \tab 7130 \tab 750 \tab 7540 \tab 7560 \tab 7790 \tab 830 \tab 8422 \tab 8440 \tab 8492 \tab 8620 \tab 9220 \tab 9260 \tab 9270 \tab 9610 \tab 9832 \tab 9891 \tab 120 \tab 1941 \tab 3211 \tab 4310 \tab 4523 \tab 5320 \tab 5995 \tab 7230 \tab 7260 \tab 7280 \tab 760 \tab 7911 \tab 810 \tab 8120 \tab 820 \tab 8200 \tab 8311 \tab 840 \tab 9570 \tab 9890 \tab 1931 \tab 7930 \tab 8510 \tab 9230 \tab 9311 \tab 9410 \tab 9811 \tab 1720 \tab 1920 \tab 340 \tab 360 \tab 7111 \tab 1350 \tab 1622 \tab 3951 \tab 4430 \tab 540 \tab 5991 \tab 3391 \tab 790 \tab 9542 \tab 9594 \tab 770 \tab 510 \tab 5322 \tab 690 \tab 7550 \tab 9560 \tab 1340 \tab 3220 \tab 5310 \tab 9210 \tab 132 \tab 1924 \tab 1730 \tab 640 \tab 3800 \tab Total \cr
   1972 \tab 1 \tab 9 \tab 3 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 2 \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 5 \tab 130 \tab 2 \tab 6 \tab 6 \tab 2 \tab 1 \tab 1 \tab 3 \tab 4 \tab 7 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 6 \tab 3 \tab 1 \tab 3 \tab 3 \tab 4 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 6 \tab 12 \tab 1 \tab 1 \tab 1 \tab 3 \tab 4 \tab 3 \tab 1 \tab 4 \tab 9 \tab 6 \tab 1 \tab 1 \tab 31 \tab 1 \tab 5 \tab 1 \tab 3 \tab 14 \tab 8 \tab 1 \tab 12 \tab 7 \tab 9 \tab 2 \tab 6 \tab 2 \tab 10 \tab 2 \tab 1 \tab 4 \tab 7 \tab 306 \tab 40 \tab 3 \tab 1 \tab 3 \tab 36 \tab 31 \tab 2 \tab 6 \tab 2 \tab 2 \tab 2 \tab 3 \tab 2 \tab 6 \tab 2 \tab 6 \tab 2 \tab 7 \tab 1 \tab 8 \tab 1 \tab 4 \tab 2 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 7 \tab 5 \tab 2 \tab 40 \tab 3 \tab 2 \tab 15 \tab 2 \tab 23 \tab 9 \tab 4 \tab 3 \tab 1 \tab 6 \tab 4 \tab 1 \tab 2 \tab 15 \tab 5 \tab 6 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 12 \tab 2 \tab 12 \tab 10 \tab 4 \tab 1 \tab 35 \tab 4 \tab 1 \tab 10 \tab 8 \tab 2 \tab 14 \tab 8 \tab 4 \tab 15 \tab 3 \tab 3 \tab 6 \tab 1 \tab 2 \tab 6 \tab 3 \tab 29 \tab 7 \tab 4 \tab 2 \tab 57 \tab 1 \tab 266 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 2 \tab 15 \tab 3 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 2 \tab 12 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab 5 \tab 129 \tab 5 \tab 5 \tab 5 \tab - \tab - \tab 1 \tab 5 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 3 \tab 3 \tab 4 \tab 8 \tab - \tab 2 \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab 2 \tab 7 \tab 5 \tab - \tab - \tab 34 \tab 1 \tab 8 \tab - \tab 2 \tab 5 \tab 7 \tab - \tab 6 \tab 6 \tab 11 \tab 1 \tab 7 \tab 1 \tab 8 \tab 1 \tab - \tab 1 \tab 5 \tab 253 \tab 33 \tab 2 \tab 2 \tab 7 \tab 56 \tab 42 \tab 3 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 7 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 5 \tab 3 \tab 1 \tab 58 \tab 1 \tab 3 \tab 16 \tab 4 \tab 11 \tab 14 \tab 5 \tab 2 \tab 1 \tab 8 \tab 1 \tab 3 \tab - \tab 9 \tab 8 \tab 6 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 3 \tab - \tab 1 \tab 13 \tab - \tab 6 \tab 7 \tab - \tab - \tab 39 \tab 3 \tab - \tab 19 \tab 7 \tab 4 \tab 11 \tab 1 \tab 5 \tab 9 \tab 1 \tab 2 \tab 6 \tab - \tab 6 \tab 7 \tab 2 \tab 25 \tab 6 \tab - \tab - \tab 61 \tab 1 \tab 183 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 41 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 5 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 6 \tab 13 \tab 2 \tab - \tab - \tab 4 \tab 3 \tab 4 \tab 2 \tab 7 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab 135 \tab 4 \tab 3 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 7 \tab 4 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 7 \tab - \tab - \tab 6 \tab 5 \tab 5 \tab - \tab - \tab - \tab 1 \tab 7 \tab - \tab 2 \tab 3 \tab 8 \tab 5 \tab 5 \tab 1 \tab 33 \tab 6 \tab 11 \tab 1 \tab - \tab 7 \tab 9 \tab 2 \tab 8 \tab - \tab 9 \tab - \tab 3 \tab 1 \tab 14 \tab 1 \tab 1 \tab 4 \tab 3 \tab 262 \tab 37 \tab 5 \tab - \tab 2 \tab 38 \tab 35 \tab 3 \tab 2 \tab 7 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab 5 \tab - \tab 3 \tab - \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 47 \tab 2 \tab 4 \tab 8 \tab 3 \tab 20 \tab 34 \tab 8 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab 15 \tab 7 \tab 9 \tab 3 \tab 2 \tab 1 \tab - \tab 2 \tab 4 \tab - \tab 2 \tab 8 \tab 3 \tab 12 \tab 7 \tab 2 \tab 1 \tab 44 \tab 2 \tab 1 \tab 10 \tab 8 \tab - \tab 11 \tab - \tab 7 \tab 14 \tab 2 \tab 2 \tab 4 \tab - \tab 5 \tab 3 \tab 4 \tab 16 \tab 16 \tab 2 \tab - \tab 44 \tab 2 \tab 201 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 9 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 3 \tab 8 \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 7 \tab 114 \tab 3 \tab 6 \tab 8 \tab 1 \tab - \tab 1 \tab 5 \tab 2 \tab 7 \tab 4 \tab 2 \tab - \tab 3 \tab - \tab - \tab 7 \tab 1 \tab - \tab 3 \tab 2 \tab 3 \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 2 \tab 4 \tab - \tab - \tab - \tab 7 \tab 5 \tab 2 \tab 2 \tab 2 \tab 9 \tab 10 \tab - \tab 2 \tab 29 \tab 4 \tab 7 \tab 2 \tab 3 \tab 10 \tab 3 \tab - \tab 11 \tab 3 \tab 10 \tab - \tab 11 \tab 1 \tab 14 \tab - \tab 1 \tab 2 \tab 7 \tab 264 \tab 15 \tab 2 \tab 2 \tab 5 \tab 40 \tab 39 \tab - \tab 1 \tab 9 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 11 \tab 1 \tab - \tab 1 \tab 5 \tab - \tab 4 \tab 4 \tab - \tab - \tab 3 \tab - \tab 8 \tab 5 \tab 4 \tab - \tab 38 \tab 1 \tab 6 \tab 23 \tab 4 \tab 23 \tab 16 \tab 14 \tab - \tab 4 \tab 9 \tab - \tab - \tab - \tab 14 \tab 12 \tab 5 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 12 \tab 1 \tab 12 \tab 9 \tab 1 \tab 2 \tab 45 \tab 3 \tab 1 \tab 23 \tab 5 \tab 1 \tab 13 \tab 2 \tab 4 \tab 16 \tab 1 \tab 1 \tab 6 \tab 2 \tab 2 \tab 3 \tab 3 \tab 31 \tab 24 \tab 1 \tab 1 \tab 40 \tab 3 \tab 176 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 6 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 8 \tab 3 \tab 2 \tab - \tab 3 \tab 5 \tab 3 \tab 2 \tab 9 \tab 2 \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 113 \tab 3 \tab 4 \tab 2 \tab - \tab - \tab - \tab 3 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 4 \tab - \tab 1 \tab 5 \tab 3 \tab - \tab - \tab - \tab 6 \tab - \tab 2 \tab - \tab 1 \tab 5 \tab 1 \tab 7 \tab 1 \tab - \tab 1 \tab - \tab 10 \tab - \tab 1 \tab 2 \tab 11 \tab 9 \tab - \tab 3 \tab 28 \tab 6 \tab 7 \tab - \tab 2 \tab 16 \tab 8 \tab 2 \tab 3 \tab 2 \tab 8 \tab 1 \tab 3 \tab - \tab 10 \tab - \tab - \tab 2 \tab 10 \tab 282 \tab 17 \tab 2 \tab 1 \tab - \tab 52 \tab 31 \tab 2 \tab 2 \tab 5 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 6 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 6 \tab 4 \tab 2 \tab 61 \tab 3 \tab 3 \tab 9 \tab 6 \tab 21 \tab 18 \tab 15 \tab 2 \tab 3 \tab 9 \tab 1 \tab 2 \tab - \tab 3 \tab 5 \tab 4 \tab 2 \tab 5 \tab - \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 11 \tab 2 \tab 11 \tab 7 \tab - \tab 1 \tab 35 \tab 1 \tab - \tab 19 \tab 15 \tab - \tab 19 \tab 3 \tab 8 \tab 9 \tab 4 \tab 1 \tab 5 \tab - \tab 3 \tab 3 \tab 1 \tab 27 \tab 19 \tab 3 \tab - \tab 58 \tab 2 \tab 209 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 3 \tab 9 \tab 10 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 5 \tab 9 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 8 \tab 117 \tab 5 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 5 \tab - \tab 6 \tab - \tab 3 \tab 1 \tab 1 \tab 7 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 7 \tab - \tab - \tab 2 \tab 9 \tab 6 \tab - \tab 2 \tab 2 \tab 4 \tab 4 \tab - \tab 3 \tab 1 \tab 9 \tab 1 \tab 3 \tab - \tab 21 \tab 4 \tab 9 \tab 1 \tab 2 \tab 14 \tab 7 \tab 3 \tab 4 \tab 2 \tab 11 \tab - \tab 6 \tab 2 \tab 21 \tab 1 \tab 1 \tab 2 \tab 4 \tab 254 \tab 39 \tab 3 \tab 2 \tab 2 \tab 36 \tab 60 \tab - \tab 3 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 8 \tab - \tab 5 \tab - \tab 5 \tab 1 \tab 2 \tab - \tab - \tab - \tab 6 \tab 9 \tab 4 \tab 1 \tab 60 \tab 3 \tab 1 \tab 13 \tab 3 \tab 19 \tab 31 \tab 10 \tab 2 \tab 1 \tab 13 \tab 3 \tab 6 \tab 1 \tab 7 \tab 19 \tab 11 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab 5 \tab - \tab - \tab 15 \tab 3 \tab 5 \tab 7 \tab 3 \tab 2 \tab 30 \tab 2 \tab - \tab 20 \tab 9 \tab 2 \tab 15 \tab 2 \tab 3 \tab 15 \tab 1 \tab 1 \tab 6 \tab - \tab 5 \tab - \tab 2 \tab 25 \tab 12 \tab 5 \tab - \tab 42 \tab 1 \tab 215 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 3 \tab 12 \tab 8 \tab - \tab - \tab 1 \tab 5 \tab 5 \tab 2 \tab 6 \tab 3 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 5 \tab 115 \tab 5 \tab 9 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 4 \tab 4 \tab - \tab 6 \tab - \tab 7 \tab - \tab - \tab 2 \tab 4 \tab - \tab 1 \tab 2 \tab 4 \tab 1 \tab 11 \tab - \tab 1 \tab 4 \tab 4 \tab 7 \tab - \tab - \tab - \tab - \tab 10 \tab 2 \tab - \tab 1 \tab 9 \tab 3 \tab - \tab - \tab 35 \tab 1 \tab 8 \tab - \tab 4 \tab 12 \tab 4 \tab 1 \tab 4 \tab 9 \tab 10 \tab - \tab 4 \tab 3 \tab 30 \tab 1 \tab - \tab 2 \tab 7 \tab 235 \tab 29 \tab - \tab 1 \tab 2 \tab 33 \tab 58 \tab 2 \tab 2 \tab 3 \tab 1 \tab 2 \tab 3 \tab - \tab 3 \tab 1 \tab 5 \tab - \tab 8 \tab 1 \tab 3 \tab - \tab 4 \tab 1 \tab 2 \tab 1 \tab 4 \tab - \tab 3 \tab 2 \tab 12 \tab 5 \tab 50 \tab 4 \tab 1 \tab 9 \tab 2 \tab 13 \tab 31 \tab 9 \tab 1 \tab 3 \tab 18 \tab 1 \tab 2 \tab - \tab 10 \tab 11 \tab 5 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 12 \tab 3 \tab 11 \tab 8 \tab 1 \tab 3 \tab 54 \tab 1 \tab 2 \tab 13 \tab 7 \tab 3 \tab 9 \tab 3 \tab 2 \tab 16 \tab 5 \tab - \tab 6 \tab 1 \tab 2 \tab 1 \tab 1 \tab 26 \tab 20 \tab 1 \tab - \tab 40 \tab 1 \tab 214 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 6 \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 3 \tab 11 \tab 9 \tab 2 \tab - \tab 9 \tab 5 \tab 1 \tab 4 \tab 5 \tab 2 \tab 1 \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 119 \tab 2 \tab 5 \tab 2 \tab - \tab 1 \tab 1 \tab 6 \tab 5 \tab 6 \tab 3 \tab 2 \tab - \tab 4 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 5 \tab - \tab 4 \tab - \tab 1 \tab 2 \tab 4 \tab 8 \tab - \tab 1 \tab 2 \tab - \tab 17 \tab 5 \tab 1 \tab 3 \tab 10 \tab 8 \tab 2 \tab - \tab 26 \tab 2 \tab 6 \tab 2 \tab - \tab 12 \tab 2 \tab 2 \tab 9 \tab 6 \tab 11 \tab 1 \tab 6 \tab - \tab 22 \tab 2 \tab - \tab 3 \tab 8 \tab 227 \tab 12 \tab - \tab - \tab 2 \tab 33 \tab 49 \tab - \tab 4 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 5 \tab - \tab 6 \tab - \tab 2 \tab 2 \tab 3 \tab - \tab 7 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 33 \tab - \tab 13 \tab 14 \tab 3 \tab 14 \tab 20 \tab 7 \tab 1 \tab - \tab 11 \tab 3 \tab 3 \tab - \tab 11 \tab 12 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab 3 \tab 13 \tab - \tab 7 \tab 9 \tab 1 \tab 2 \tab 39 \tab 1 \tab - \tab 12 \tab 4 \tab 2 \tab 9 \tab 4 \tab 10 \tab 9 \tab 4 \tab - \tab 4 \tab - \tab 2 \tab 5 \tab 3 \tab 22 \tab 14 \tab 1 \tab - \tab 48 \tab 1 \tab 238 \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 5 \tab 5 \tab - \tab - \tab 2 \tab 10 \tab 2 \tab 4 \tab 13 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 6 \tab 111 \tab 6 \tab 4 \tab 3 \tab 1 \tab - \tab 1 \tab 10 \tab 3 \tab 5 \tab 3 \tab 5 \tab - \tab 3 \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 10 \tab - \tab - \tab 4 \tab 10 \tab 5 \tab - \tab 2 \tab 1 \tab 2 \tab 13 \tab - \tab 5 \tab 2 \tab 11 \tab 8 \tab 1 \tab 1 \tab 48 \tab 2 \tab 10 \tab 2 \tab 3 \tab 17 \tab 15 \tab 3 \tab 8 \tab 9 \tab 6 \tab - \tab 11 \tab - \tab 21 \tab 1 \tab - \tab 2 \tab 5 \tab 264 \tab 37 \tab 3 \tab 3 \tab 2 \tab 33 \tab 53 \tab 2 \tab 1 \tab 10 \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab 10 \tab - \tab 5 \tab - \tab 6 \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 4 \tab 4 \tab 1 \tab 72 \tab 1 \tab 11 \tab 8 \tab 3 \tab 21 \tab 22 \tab 13 \tab 3 \tab 1 \tab 12 \tab 4 \tab 6 \tab 1 \tab 16 \tab 10 \tab 9 \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 11 \tab - \tab 12 \tab 18 \tab 3 \tab 2 \tab 48 \tab 3 \tab 1 \tab 35 \tab - \tab 1 \tab 18 \tab 2 \tab 4 \tab 15 \tab 1 \tab 2 \tab 4 \tab 2 \tab 7 \tab 4 \tab 10 \tab 48 \tab 12 \tab - \tab - \tab 58 \tab 5 \tab 357 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 3 \tab 14 \tab 10 \tab - \tab - \tab 13 \tab 1 \tab 6 \tab 2 \tab 12 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 12 \tab 143 \tab 6 \tab 6 \tab 10 \tab - \tab - \tab - \tab 5 \tab 7 \tab 4 \tab 6 \tab - \tab - \tab 3 \tab - \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab 12 \tab - \tab - \tab 10 \tab 8 \tab 5 \tab - \tab 2 \tab 1 \tab 2 \tab 7 \tab 3 \tab 4 \tab 1 \tab 7 \tab 3 \tab 3 \tab 1 \tab 41 \tab - \tab 2 \tab - \tab 1 \tab 24 \tab 2 \tab 3 \tab 10 \tab 5 \tab 10 \tab - \tab 7 \tab - \tab 16 \tab - \tab - \tab 3 \tab 8 \tab 191 \tab 20 \tab 6 \tab 2 \tab 1 \tab 53 \tab 25 \tab - \tab 7 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 9 \tab 1 \tab 2 \tab - \tab 5 \tab 1 \tab 6 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 10 \tab 6 \tab 2 \tab - \tab 40 \tab 6 \tab 15 \tab 17 \tab 2 \tab 20 \tab 20 \tab 9 \tab 3 \tab 2 \tab 16 \tab 2 \tab 3 \tab - \tab 9 \tab 12 \tab 10 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 13 \tab 3 \tab 7 \tab 15 \tab 3 \tab 3 \tab 49 \tab 2 \tab - \tab 10 \tab 4 \tab 3 \tab 8 \tab 2 \tab 8 \tab 13 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 34 \tab 16 \tab 2 \tab 1 \tab 62 \tab 3 \tab 222 \tab - \tab - \tab 1 \tab - \tab - \tab 5 \tab 1 \tab 4 \tab - \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 5 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 3 \tab 5 \tab 5 \tab - \tab 1 \tab 5 \tab 4 \tab 3 \tab 3 \tab 5 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 3 \tab 159 \tab 4 \tab 8 \tab 2 \tab 2 \tab - \tab - \tab 3 \tab 9 \tab 6 \tab 3 \tab 6 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 6 \tab - \tab 1 \tab 6 \tab 8 \tab 5 \tab - \tab 1 \tab - \tab 3 \tab 8 \tab 2 \tab 1 \tab - \tab 12 \tab 3 \tab 1 \tab 1 \tab 32 \tab 1 \tab 6 \tab 1 \tab 1 \tab 9 \tab 10 \tab 1 \tab 7 \tab 5 \tab 8 \tab - \tab 6 \tab 1 \tab 25 \tab - \tab - \tab 1 \tab 3 \tab 231 \tab 23 \tab 1 \tab 5 \tab 1 \tab 27 \tab 26 \tab 2 \tab 3 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 8 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab 4 \tab 1 \tab 2 \tab 39 \tab 4 \tab 19 \tab 6 \tab 1 \tab 8 \tab 15 \tab 6 \tab - \tab - \tab 16 \tab 4 \tab 4 \tab 1 \tab 15 \tab 9 \tab 4 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 9 \tab 1 \tab 11 \tab 10 \tab 3 \tab 1 \tab 21 \tab - \tab 1 \tab 21 \tab 5 \tab 3 \tab 4 \tab 3 \tab 3 \tab 10 \tab 7 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 7 \tab 33 \tab 7 \tab 3 \tab - \tab 67 \tab 1 \tab 224 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 8 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 6 \tab 9 \tab 8 \tab - \tab - \tab 5 \tab 6 \tab 1 \tab 4 \tab 14 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 10 \tab 147 \tab 6 \tab 5 \tab 10 \tab 6 \tab 1 \tab 1 \tab 5 \tab 4 \tab 10 \tab 3 \tab - \tab - \tab 6 \tab 2 \tab 1 \tab 6 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab 10 \tab - \tab - \tab 3 \tab 6 \tab 8 \tab - \tab - \tab 1 \tab 1 \tab 11 \tab - \tab 6 \tab 1 \tab 14 \tab 7 \tab 2 \tab - \tab 29 \tab 1 \tab 9 \tab 1 \tab 2 \tab 12 \tab 5 \tab 2 \tab 5 \tab 6 \tab 12 \tab - \tab 4 \tab - \tab 16 \tab - \tab 2 \tab 2 \tab 9 \tab 223 \tab 23 \tab 3 \tab 1 \tab 2 \tab 63 \tab 17 \tab 1 \tab 6 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 13 \tab - \tab 9 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 5 \tab 3 \tab 8 \tab - \tab 45 \tab 3 \tab 11 \tab 12 \tab 7 \tab 23 \tab 22 \tab 4 \tab - \tab 3 \tab 16 \tab 1 \tab 2 \tab - \tab 11 \tab 12 \tab 5 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 6 \tab - \tab 7 \tab 18 \tab 2 \tab 2 \tab 32 \tab - \tab - \tab 25 \tab 2 \tab 3 \tab 7 \tab 4 \tab 4 \tab 19 \tab 4 \tab 2 \tab 7 \tab 1 \tab 2 \tab 2 \tab - \tab 38 \tab 4 \tab 2 \tab - \tab 35 \tab 2 \tab 203 \tab - \tab - \tab 1 \tab - \tab - \tab 6 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 5 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 2 \tab 4 \tab 7 \tab 2 \tab - \tab 7 \tab 2 \tab 1 \tab 3 \tab 8 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab 3 \tab 155 \tab 4 \tab 8 \tab 4 \tab 2 \tab - \tab 2 \tab 3 \tab 2 \tab 7 \tab 2 \tab 3 \tab 1 \tab 2 \tab 4 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 7 \tab 1 \tab - \tab 4 \tab 9 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 13 \tab 1 \tab 3 \tab - \tab 6 \tab 8 \tab 2 \tab 3 \tab 55 \tab - \tab 2 \tab 4 \tab 4 \tab 11 \tab 2 \tab - \tab 2 \tab 5 \tab 12 \tab - \tab 5 \tab 3 \tab 13 \tab - \tab - \tab 2 \tab 5 \tab 213 \tab 23 \tab 3 \tab 1 \tab 2 \tab 47 \tab 15 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 6 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 6 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 5 \tab 2 \tab 5 \tab 1 \tab 41 \tab 3 \tab 7 \tab 26 \tab 1 \tab 7 \tab 17 \tab 3 \tab - \tab - \tab 15 \tab 2 \tab 3 \tab - \tab 10 \tab 18 \tab 6 \tab 2 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 11 \tab 1 \tab 8 \tab 3 \tab 4 \tab 2 \tab 33 \tab - \tab - \tab 8 \tab - \tab - \tab 5 \tab 2 \tab 5 \tab 8 \tab 6 \tab 3 \tab 5 \tab 1 \tab 6 \tab 5 \tab 4 \tab 32 \tab 1 \tab 1 \tab - \tab 53 \tab 1 \tab 244 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 6 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 2 \tab 19 \tab 9 \tab - \tab - \tab 9 \tab 11 \tab 2 \tab 2 \tab 11 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 5 \tab 135 \tab 3 \tab 11 \tab 7 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 11 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 2 \tab 2 \tab 9 \tab - \tab 6 \tab 2 \tab - \tab 4 \tab 5 \tab 12 \tab - \tab - \tab - \tab 1 \tab 9 \tab 2 \tab 1 \tab 1 \tab 10 \tab 5 \tab - \tab 1 \tab 35 \tab 2 \tab 5 \tab 1 \tab 2 \tab 14 \tab 10 \tab 3 \tab 2 \tab 8 \tab 9 \tab - \tab 6 \tab 1 \tab 25 \tab - \tab - \tab 1 \tab 11 \tab 270 \tab 38 \tab 2 \tab 2 \tab 2 \tab 41 \tab 34 \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 7 \tab - \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 2 \tab 6 \tab 1 \tab 3 \tab 5 \tab 1 \tab 88 \tab 1 \tab 12 \tab 8 \tab 1 \tab 15 \tab 14 \tab 9 \tab 1 \tab 3 \tab 14 \tab - \tab 2 \tab - \tab 12 \tab 19 \tab 3 \tab - \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 11 \tab 1 \tab 11 \tab 12 \tab - \tab 4 \tab 44 \tab 1 \tab - \tab 44 \tab 1 \tab - \tab 9 \tab 3 \tab 8 \tab 5 \tab 4 \tab 1 \tab 1 \tab 1 \tab 5 \tab 5 \tab 4 \tab 41 \tab 20 \tab 6 \tab 1 \tab 64 \tab 3 \tab 315 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 3 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab 3 \tab - \tab 2 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 3 \tab 14 \tab 5 \tab - \tab 1 \tab 8 \tab 7 \tab 2 \tab 7 \tab 10 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 2 \tab 160 \tab 5 \tab 8 \tab 9 \tab 6 \tab - \tab - \tab 4 \tab 2 \tab 16 \tab 2 \tab 5 \tab - \tab 4 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 7 \tab - \tab - \tab 5 \tab 7 \tab 6 \tab - \tab 1 \tab - \tab - \tab 10 \tab 1 \tab 5 \tab 1 \tab 7 \tab 1 \tab - \tab 2 \tab 45 \tab 3 \tab 2 \tab 1 \tab - \tab 8 \tab 13 \tab 2 \tab 4 \tab 5 \tab 9 \tab - \tab 7 \tab 2 \tab 20 \tab - \tab - \tab - \tab 7 \tab 194 \tab 32 \tab 3 \tab 1 \tab 1 \tab 31 \tab 16 \tab 1 \tab 3 \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab 4 \tab 2 \tab 5 \tab - \tab 5 \tab - \tab 4 \tab - \tab 3 \tab - \tab - \tab - \tab 4 \tab - \tab 4 \tab - \tab 7 \tab - \tab 39 \tab 1 \tab 1 \tab 22 \tab 1 \tab 20 \tab 18 \tab 7 \tab 1 \tab 2 \tab 13 \tab 3 \tab 3 \tab - \tab 7 \tab 12 \tab 5 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab 7 \tab 1 \tab 8 \tab 5 \tab 1 \tab 4 \tab 32 \tab 2 \tab - \tab 21 \tab 3 \tab - \tab 7 \tab 3 \tab 1 \tab 15 \tab 4 \tab 2 \tab 1 \tab 1 \tab 6 \tab 9 \tab 7 \tab 39 \tab 7 \tab 2 \tab - \tab 43 \tab 5 \tab 222 \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab 2 \tab 5 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 3 \tab 10 \tab 5 \tab 1 \tab - \tab 12 \tab 9 \tab 5 \tab 6 \tab 12 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 7 \tab 160 \tab 9 \tab 7 \tab 6 \tab 3 \tab 2 \tab 1 \tab 6 \tab 5 \tab 6 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 6 \tab 1 \tab - \tab 1 \tab - \tab 6 \tab 2 \tab 12 \tab - \tab - \tab 5 \tab 5 \tab 2 \tab - \tab 2 \tab 3 \tab 3 \tab 11 \tab 2 \tab 5 \tab 3 \tab 5 \tab 12 \tab 3 \tab 6 \tab 29 \tab - \tab 6 \tab 3 \tab 1 \tab 9 \tab 5 \tab 4 \tab 5 \tab 6 \tab 11 \tab - \tab 6 \tab - \tab 16 \tab - \tab - \tab 1 \tab 7 \tab 200 \tab 16 \tab - \tab 1 \tab 3 \tab 38 \tab 28 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 10 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 6 \tab 1 \tab 48 \tab 4 \tab 5 \tab 15 \tab 5 \tab 27 \tab 21 \tab 11 \tab 1 \tab 2 \tab 16 \tab 1 \tab 1 \tab - \tab 9 \tab 10 \tab 6 \tab 1 \tab 7 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 5 \tab 1 \tab 6 \tab 10 \tab 2 \tab 1 \tab 24 \tab - \tab 1 \tab 30 \tab 4 \tab 6 \tab 9 \tab 5 \tab 2 \tab 15 \tab 2 \tab - \tab 4 \tab - \tab 4 \tab 2 \tab 5 \tab 45 \tab 8 \tab - \tab - \tab 32 \tab 2 \tab 242 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 6 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1537 \cr
   1990 \tab 5 \tab 9 \tab 1 \tab - \tab - \tab 6 \tab 6 \tab 3 \tab 1 \tab 8 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 144 \tab 8 \tab 9 \tab 10 \tab 1 \tab - \tab - \tab 8 \tab 3 \tab 12 \tab 1 \tab 2 \tab - \tab 4 \tab - \tab - \tab 6 \tab 3 \tab - \tab 1 \tab - \tab 4 \tab 2 \tab 7 \tab - \tab 1 \tab 2 \tab 7 \tab 7 \tab - \tab - \tab 1 \tab - \tab 17 \tab - \tab 5 \tab 1 \tab 8 \tab 5 \tab - \tab 1 \tab 38 \tab 1 \tab 4 \tab 4 \tab 1 \tab 11 \tab 6 \tab 3 \tab 2 \tab 3 \tab 6 \tab - \tab 3 \tab - \tab 14 \tab 1 \tab 1 \tab - \tab 6 \tab 163 \tab 17 \tab 2 \tab - \tab - \tab 39 \tab 14 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 7 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 4 \tab 1 \tab 9 \tab - \tab 34 \tab 3 \tab 4 \tab 11 \tab 2 \tab 20 \tab 16 \tab 11 \tab - \tab 1 \tab 8 \tab 2 \tab 3 \tab 1 \tab 9 \tab 19 \tab 4 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 10 \tab 2 \tab 10 \tab 3 \tab 1 \tab 2 \tab 28 \tab 1 \tab - \tab 26 \tab 2 \tab 3 \tab 5 \tab 1 \tab 7 \tab 9 \tab 2 \tab - \tab 4 \tab - \tab 5 \tab 1 \tab 2 \tab 37 \tab 8 \tab - \tab - \tab 47 \tab - \tab 234 \tab - \tab - \tab 3 \tab - \tab 1 \tab 4 \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 45 \tab 175 \tab 94 \tab 8 \tab 3 \tab 94 \tab 84 \tab 44 \tab 54 \tab 159 \tab 31 \tab 10 \tab 20 \tab 14 \tab 22 \tab 14 \tab 8 \tab 10 \tab 26 \tab 91 \tab 2286 \tab 80 \tab 105 \tab 90 \tab 31 \tab 6 \tab 9 \tab 75 \tab 58 \tab 115 \tab 36 \tab 48 \tab 3 \tab 56 \tab 11 \tab 6 \tab 74 \tab 34 \tab 6 \tab 24 \tab 14 \tab 54 \tab 23 \tab 120 \tab 4 \tab 10 \tab 69 \tab 100 \tab 109 \tab 2 \tab 15 \tab 18 \tab 29 \tab 159 \tab 23 \tab 46 \tab 28 \tab 152 \tab 99 \tab 23 \tab 23 \tab 589 \tab 35 \tab 107 \tab 24 \tab 31 \tab 205 \tab 116 \tab 32 \tab 102 \tab 87 \tab 162 \tab 5 \tab 101 \tab 17 \tab 295 \tab 10 \tab 7 \tab 32 \tab 112 \tab 4032 \tab 451 \tab 40 \tab 25 \tab 37 \tab 696 \tab 573 \tab 28 \tab 47 \tab 53 \tab 14 \tab 22 \tab 18 \tab 13 \tab 42 \tab 10 \tab 112 \tab 9 \tab 67 \tab 9 \tab 68 \tab 4 \tab 64 \tab 22 \tab 24 \tab 10 \tab 30 \tab 17 \tab 54 \tab 65 \tab 82 \tab 20 \tab 833 \tab 43 \tab 118 \tab 232 \tab 50 \tab 305 \tab 338 \tab 145 \tab 21 \tab 28 \tab 203 \tab 35 \tab 45 \tab 6 \tab 182 \tab 200 \tab 100 \tab 21 \tab 33 \tab 22 \tab 4 \tab 8 \tab 47 \tab 6 \tab 14 \tab 179 \tab 24 \tab 156 \tab 158 \tab 31 \tab 33 \tab 632 \tab 26 \tab 8 \tab 346 \tab 84 \tab 33 \tab 173 \tab 48 \tab 85 \tab 212 \tab 53 \tab 22 \tab 72 \tab 12 \tab 63 \tab 59 \tab 64 \tab 548 \tab 201 \tab 33 \tab 5 \tab 851 \tab 34 \tab 50090 \tab 4 \tab 8 \tab 18 \tab 4 \tab 11 \tab 48 \tab 20 \tab 33 \tab 12 \tab 26 \tab 5 \tab 28 \tab 2 \tab 7 \tab 9 \tab 15 \tab 7 \tab 2 \tab 5 \tab 8 \tab 9 \tab 66 \tab 9 \tab 14 \tab 12 \tab 3 \tab 35 \tab 8 \tab 14 \tab 9 \tab 1 \tab 12 \tab 6 \tab 41 \tab 10 \tab 5 \tab 42 \tab 7 \tab 3 \tab 11 \tab 15 \tab 6 \tab 5 \tab 6 \tab 7 \tab 3 \tab 8 \tab 15 \tab 2 \tab 3 \tab 5 \tab 7 \tab 4 \tab 3 \tab 3 \tab 17 \tab 3 \tab 5 \tab 10 \tab 29 \tab 9 \tab 3 \tab 6 \tab 1 \tab 2 \tab 1 \tab 8 \tab 4 \tab 3 \tab 2 \tab 2 \tab 29 \tab 15 \tab 5 \tab 4 \tab 9 \tab 3 \tab 1 \tab 8 \tab 4 \tab 2 \tab 5 \tab 2 \tab 9 \tab 2 \tab 4 \tab 4 \tab 1 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
