% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_596.R
\name{paisco88}
\alias{paisco88}
\title{R's father's occupation, 1980 census & 1988 isco code}
\source{
General Social Survey https://gss.norc.org
}
\description{
paisco88
}
\details{
Question 1643. Respondent's Father's Occupation, 1988 Census
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 110 \tab 1210 \tab 1220 \tab 1221 \tab 1223 \tab 1225 \tab 1226 \tab 1227 \tab 1229 \tab 1231 \tab 1233 \tab 1234 \tab 1235 \tab 1239 \tab 1311 \tab 1313 \tab 1315 \tab 1319 \tab 2110 \tab 2113 \tab 2141 \tab 2142 \tab 2143 \tab 2144 \tab 2145 \tab 2146 \tab 2147 \tab 2148 \tab 2149 \tab 2211 \tab 2213 \tab 2221 \tab 2222 \tab 2224 \tab 2310 \tab 2320 \tab 2411 \tab 2412 \tab 2421 \tab 2422 \tab 2441 \tab 2453 \tab 2460 \tab 3118 \tab 3122 \tab 3131 \tab 3142 \tab 3143 \tab 3152 \tab 3221 \tab 3224 \tab 3226 \tab 3231 \tab 3411 \tab 3412 \tab 3413 \tab 3415 \tab 3416 \tab 3417 \tab 3432 \tab 3434 \tab 3450 \tab 3471 \tab 4110 \tab 4111 \tab 4114 \tab 4121 \tab 4122 \tab 4131 \tab 4132 \tab 4133 \tab 4142 \tab 4143 \tab 4190 \tab 4212 \tab 4213 \tab 4223 \tab 5111 \tab 5112 \tab 5122 \tab 5123 \tab 5141 \tab 5143 \tab 5161 \tab 5162 \tab 5169 \tab 5200 \tab 5220 \tab 6113 \tab 6141 \tab 7112 \tab 7122 \tab 7123 \tab 7124 \tab 7131 \tab 7132 \tab 7133 \tab 7136 \tab 7137 \tab 7141 \tab 7200 \tab 7212 \tab 7213 \tab 7214 \tab 7222 \tab 7230 \tab 7231 \tab 7233 \tab 7240 \tab 7241 \tab 7243 \tab 7245 \tab 7300 \tab 7313 \tab 7324 \tab 7341 \tab 7343 \tab 7411 \tab 7412 \tab 7415 \tab 7433 \tab 7435 \tab 7437 \tab 7442 \tab 7510 \tab 8111 \tab 8122 \tab 8141 \tab 8151 \tab 8152 \tab 8162 \tab 8200 \tab 8221 \tab 8230 \tab 8231 \tab 8232 \tab 8251 \tab 8261 \tab 8263 \tab 8264 \tab 8266 \tab 8269 \tab 8270 \tab 8282 \tab 8290 \tab 8311 \tab 8312 \tab 8322 \tab 8323 \tab 8332 \tab 8333 \tab 8334 \tab 8340 \tab 9110 \tab 9131 \tab 9140 \tab 9141 \tab 9151 \tab 9153 \tab 9161 \tab 9211 \tab 9212 \tab 9213 \tab 9311 \tab 9312 \tab 9313 \tab 9320 \tab 9322 \tab 9333 \tab iap \tab uncodeable \tab 2223 \tab 2331 \tab 2419 \tab 2446 \tab 2451 \tab 2452 \tab 2455 \tab 3111 \tab 3112 \tab 3132 \tab 3144 \tab 3151 \tab 3211 \tab 3410 \tab 3422 \tab 3429 \tab 3475 \tab 4115 \tab 4211 \tab 4222 \tab 5163 \tab 7129 \tab 7135 \tab 7223 \tab 7232 \tab 7311 \tab 7420 \tab 7422 \tab 8113 \tab 8161 \tab 8253 \tab 8262 \tab 8275 \tab 8310 \tab 9142 \tab 9321 \tab 1232 \tab 2112 \tab 2122 \tab 2131 \tab 2359 \tab 3340 \tab 3442 \tab 3443 \tab 3460 \tab 4215 \tab 4221 \tab 5100 \tab 6129 \tab 7215 \tab 8100 \tab 8123 \tab 8155 \tab 8163 \tab 8212 \tab 9132 \tab 9152 \tab 2132 \tab 3472 \tab 7224 \tab 2442 \tab 3212 \tab 3470 \tab 5120 \tab 5131 \tab 7111 \tab 1110 \tab 3110 \tab 3141 \tab 2121 \tab 2230 \tab 3115 \tab 4100 \tab 4141 \tab 5113 \tab 1236 \tab 8224 \tab 3133 \tab 3480 \tab 2229 \tab 2340 \tab 3223 \tab 7134 \tab 7143 \tab 2444 \tab 3400 \tab 3229 \tab 4113 \tab 8223 \tab 2445 \tab 2454 \tab 2111 \tab 2212 \tab 2332 \tab 5139 \tab 3227 \tab 2431 \tab 5133 \tab 6124 \tab 1237 \tab 6154 \tab 9160 \tab 2432 \tab 3121 \tab 5132 \tab 9113 \tab 3300 \tab 2440 \tab 7216 \tab Total \cr
   1972 \tab 9 \tab 3 \tab 8 \tab 300 \tab 11 \tab 5 \tab 6 \tab 3 \tab 2 \tab 7 \tab 21 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 3 \tab 14 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 7 \tab 2 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 8 \tab 3 \tab 3 \tab 1 \tab 6 \tab 9 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 11 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 13 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 9 \tab 6 \tab 12 \tab 8 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 4 \tab 1 \tab 9 \tab 2 \tab 2 \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 6 \tab 3 \tab 12 \tab 2 \tab 8 \tab 12 \tab 5 \tab 1 \tab 12 \tab 6 \tab 13 \tab 1 \tab 9 \tab 3 \tab 34 \tab 1 \tab 1 \tab 4 \tab 15 \tab 5 \tab 12 \tab 10 \tab 5 \tab 4 \tab 1 \tab 20 \tab 4 \tab 24 \tab 2 \tab 2 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 7 \tab 7 \tab 2 \tab 8 \tab 1 \tab 2 \tab 4 \tab 25 \tab 30 \tab 7 \tab 3 \tab 3 \tab 1 \tab 11 \tab 7 \tab 1 \tab 2 \tab 1 \tab 7 \tab 3 \tab 3 \tab 2 \tab 4 \tab 2 \tab 7 \tab 1 \tab 2 \tab 9 \tab 6 \tab 3 \tab 44 \tab 3 \tab 9 \tab 5 \tab 3 \tab 3 \tab 2 \tab 2 \tab 17 \tab 1 \tab 1 \tab 1 \tab 1 \tab 54 \tab 1 \tab 1 \tab 1 \tab 15 \tab 1 \tab 1 \tab 4 \tab 21 \tab 255 \tab 151 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 8 \tab 7 \tab 11 \tab 233 \tab 23 \tab 4 \tab 8 \tab 6 \tab 3 \tab 4 \tab 41 \tab 1 \tab 2 \tab 1 \tab 1 \tab 7 \tab 2 \tab 14 \tab - \tab 1 \tab 4 \tab 5 \tab 4 \tab 2 \tab 6 \tab - \tab 2 \tab - \tab 6 \tab - \tab - \tab 4 \tab 2 \tab 7 \tab 3 \tab 2 \tab 12 \tab - \tab 3 \tab - \tab - \tab - \tab 13 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 11 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 8 \tab 4 \tab 21 \tab 6 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab 4 \tab - \tab 4 \tab 1 \tab 1 \tab 7 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 6 \tab 5 \tab 6 \tab 1 \tab 4 \tab 11 \tab 6 \tab 1 \tab 21 \tab 6 \tab 7 \tab 1 \tab 10 \tab 1 \tab 44 \tab - \tab - \tab 4 \tab 8 \tab 9 \tab 12 \tab 13 \tab 7 \tab 4 \tab 10 \tab 20 \tab 11 \tab 17 \tab 4 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 5 \tab 1 \tab - \tab 4 \tab - \tab - \tab 2 \tab 33 \tab 34 \tab - \tab 6 \tab 4 \tab 3 \tab 9 \tab 26 \tab 1 \tab 3 \tab - \tab 3 \tab 10 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 4 \tab 5 \tab 45 \tab 1 \tab 11 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 9 \tab - \tab 1 \tab 2 \tab - \tab 81 \tab 1 \tab 2 \tab 7 \tab 25 \tab 1 \tab 1 \tab 2 \tab 21 \tab 179 \tab 75 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab 2 \tab 3 \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 14 \tab 8 \tab 4 \tab 231 \tab 13 \tab 9 \tab 7 \tab 4 \tab 2 \tab 4 \tab 37 \tab 3 \tab 2 \tab 2 \tab 5 \tab 4 \tab - \tab 13 \tab - \tab 4 \tab 1 \tab 5 \tab 5 \tab 2 \tab 4 \tab 4 \tab 2 \tab - \tab 4 \tab - \tab 1 \tab 3 \tab 5 \tab 4 \tab 4 \tab 1 \tab 12 \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 6 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 11 \tab - \tab - \tab - \tab 3 \tab 4 \tab 8 \tab 5 \tab 19 \tab 6 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 6 \tab - \tab 9 \tab 2 \tab 2 \tab 15 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 5 \tab 3 \tab 9 \tab 1 \tab - \tab 9 \tab 5 \tab 1 \tab 17 \tab 6 \tab 4 \tab 1 \tab 9 \tab - \tab 44 \tab 1 \tab 1 \tab 2 \tab 18 \tab 4 \tab 10 \tab 6 \tab 7 \tab 5 \tab 5 \tab 10 \tab 9 \tab 28 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 2 \tab - \tab 3 \tab 4 \tab - \tab 5 \tab 5 \tab - \tab 2 \tab 2 \tab - \tab - \tab 22 \tab 25 \tab 4 \tab 3 \tab 2 \tab 1 \tab 10 \tab 19 \tab 1 \tab 6 \tab 1 \tab 9 \tab 6 \tab - \tab 4 \tab 4 \tab - \tab 2 \tab - \tab 3 \tab 3 \tab 4 \tab 6 \tab 36 \tab 4 \tab 8 \tab 6 \tab 2 \tab 1 \tab 7 \tab 4 \tab 10 \tab - \tab 1 \tab - \tab 2 \tab 63 \tab - \tab - \tab 11 \tab 20 \tab - \tab - \tab - \tab 17 \tab 248 \tab 46 \tab - \tab 4 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 4 \tab 2 \tab 2 \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 15 \tab 4 \tab 7 \tab 261 \tab 17 \tab 8 \tab 7 \tab 4 \tab 3 \tab 9 \tab 37 \tab - \tab - \tab 1 \tab 4 \tab 6 \tab 1 \tab 20 \tab - \tab - \tab 2 \tab 2 \tab 5 \tab 2 \tab 6 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 7 \tab 2 \tab 5 \tab 3 \tab 2 \tab 7 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 8 \tab 1 \tab - \tab - \tab 6 \tab - \tab 11 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 6 \tab 10 \tab 7 \tab 3 \tab 2 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab - \tab 4 \tab - \tab - \tab 11 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 5 \tab 11 \tab - \tab 4 \tab 10 \tab 7 \tab 1 \tab 16 \tab 4 \tab 10 \tab 3 \tab 7 \tab - \tab 46 \tab 2 \tab 3 \tab 3 \tab 14 \tab 8 \tab 13 \tab 7 \tab 12 \tab 3 \tab - \tab 20 \tab 7 \tab 26 \tab 4 \tab 2 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 8 \tab 1 \tab - \tab 5 \tab - \tab - \tab 2 \tab 26 \tab 26 \tab - \tab 6 \tab 2 \tab 1 \tab 11 \tab 13 \tab 1 \tab 9 \tab 1 \tab 5 \tab 5 \tab 1 \tab 6 \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 8 \tab 6 \tab 3 \tab 60 \tab 5 \tab 7 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 11 \tab - \tab 1 \tab - \tab 3 \tab 25 \tab 1 \tab 4 \tab 10 \tab 25 \tab - \tab - \tab 2 \tab 21 \tab 172 \tab 117 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 3 \tab 4 \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 10 \tab 8 \tab 7 \tab 250 \tab 21 \tab 6 \tab 7 \tab 7 \tab 2 \tab 4 \tab 53 \tab - \tab 3 \tab - \tab 3 \tab 4 \tab - \tab 15 \tab - \tab 2 \tab 1 \tab 3 \tab 5 \tab 1 \tab 5 \tab 2 \tab 1 \tab - \tab 6 \tab - \tab - \tab 10 \tab 2 \tab - \tab 3 \tab 5 \tab 7 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 9 \tab 3 \tab - \tab 4 \tab - \tab - \tab 8 \tab - \tab - \tab - \tab 1 \tab - \tab 11 \tab 9 \tab 18 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 3 \tab - \tab 7 \tab 2 \tab - \tab 5 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab 8 \tab 8 \tab 4 \tab 3 \tab - \tab 4 \tab 9 \tab 5 \tab 1 \tab 9 \tab 5 \tab 9 \tab 2 \tab 6 \tab - \tab 36 \tab 1 \tab - \tab - \tab 4 \tab 8 \tab 12 \tab 14 \tab 5 \tab 3 \tab 9 \tab 23 \tab 11 \tab 31 \tab 5 \tab 2 \tab - \tab 5 \tab 1 \tab - \tab - \tab 6 \tab 3 \tab - \tab 3 \tab 2 \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 30 \tab 27 \tab 4 \tab 4 \tab 6 \tab 2 \tab 12 \tab 24 \tab 2 \tab 6 \tab - \tab 9 \tab 6 \tab - \tab 3 \tab 4 \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab 4 \tab 2 \tab 49 \tab 1 \tab 9 \tab 9 \tab 4 \tab - \tab 7 \tab 1 \tab 16 \tab 1 \tab 1 \tab 1 \tab 2 \tab 42 \tab - \tab 1 \tab 7 \tab 30 \tab 1 \tab - \tab 1 \tab 32 \tab 211 \tab 51 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 21 \tab 3 \tab 4 \tab 250 \tab 8 \tab 7 \tab 2 \tab 4 \tab 4 \tab 6 \tab 22 \tab 1 \tab 3 \tab 1 \tab 2 \tab 5 \tab 2 \tab 9 \tab - \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 5 \tab - \tab 1 \tab 4 \tab 3 \tab 1 \tab 3 \tab 1 \tab 7 \tab - \tab 8 \tab - \tab - \tab 2 \tab 8 \tab 5 \tab - \tab - \tab 4 \tab 2 \tab 5 \tab - \tab 1 \tab - \tab - \tab 3 \tab 7 \tab 1 \tab 5 \tab 3 \tab - \tab - \tab - \tab 1 \tab 3 \tab 3 \tab - \tab - \tab 5 \tab 1 \tab 8 \tab 2 \tab - \tab 14 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab 5 \tab 3 \tab 1 \tab 2 \tab 12 \tab 5 \tab 1 \tab 13 \tab 9 \tab 11 \tab 7 \tab 7 \tab 3 \tab 30 \tab 3 \tab - \tab 2 \tab 8 \tab 11 \tab 11 \tab 8 \tab 18 \tab 7 \tab 4 \tab 22 \tab 9 \tab 26 \tab 4 \tab 1 \tab - \tab 5 \tab 5 \tab 1 \tab 1 \tab 6 \tab 1 \tab - \tab 2 \tab 8 \tab - \tab 5 \tab - \tab 3 \tab 1 \tab 32 \tab 45 \tab 4 \tab 5 \tab 1 \tab - \tab 9 \tab 13 \tab 1 \tab 6 \tab 1 \tab 9 \tab 4 \tab 1 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 12 \tab 5 \tab 3 \tab 36 \tab 2 \tab 7 \tab 3 \tab - \tab 1 \tab 4 \tab 1 \tab 13 \tab - \tab 1 \tab 2 \tab 1 \tab 43 \tab - \tab 2 \tab 14 \tab 18 \tab - \tab - \tab 1 \tab 30 \tab 223 \tab 142 \tab 3 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 6 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 30 \tab 5 \tab 13 \tab 216 \tab 22 \tab 8 \tab 3 \tab 6 \tab 1 \tab 5 \tab 51 \tab 1 \tab 5 \tab 1 \tab 2 \tab 6 \tab 1 \tab 16 \tab - \tab 3 \tab 1 \tab 4 \tab 6 \tab - \tab 7 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 7 \tab - \tab 3 \tab 2 \tab 4 \tab 12 \tab 1 \tab 9 \tab - \tab - \tab 3 \tab 6 \tab 4 \tab - \tab - \tab - \tab - \tab 9 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 9 \tab 2 \tab 27 \tab 4 \tab - \tab - \tab - \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab 6 \tab - \tab 7 \tab - \tab 2 \tab 13 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 4 \tab 2 \tab 7 \tab 4 \tab 1 \tab 9 \tab 14 \tab - \tab 1 \tab 15 \tab 4 \tab 16 \tab 4 \tab 10 \tab 2 \tab 56 \tab 3 \tab - \tab 1 \tab 8 \tab 13 \tab 11 \tab 14 \tab 11 \tab 2 \tab 9 \tab 31 \tab 21 \tab 21 \tab 2 \tab 5 \tab 2 \tab 4 \tab 3 \tab 3 \tab - \tab 4 \tab 1 \tab 2 \tab 3 \tab 6 \tab 1 \tab 5 \tab 1 \tab 1 \tab 3 \tab 29 \tab 39 \tab 2 \tab 3 \tab 4 \tab - \tab 8 \tab 26 \tab 2 \tab 7 \tab 1 \tab 8 \tab 8 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 9 \tab 7 \tab 2 \tab 42 \tab - \tab 11 \tab 4 \tab 3 \tab - \tab - \tab - \tab 13 \tab 1 \tab - \tab - \tab 1 \tab 41 \tab 1 \tab 1 \tab 15 \tab 15 \tab 1 \tab 1 \tab 1 \tab 23 \tab 220 \tab - \tab - \tab 5 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 3 \tab 4 \tab - \tab 5 \tab 2 \tab - \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 23 \tab 4 \tab 11 \tab 203 \tab 19 \tab 8 \tab 4 \tab 10 \tab 5 \tab 7 \tab 63 \tab 2 \tab - \tab 3 \tab 2 \tab 6 \tab 5 \tab 15 \tab - \tab 3 \tab 3 \tab 1 \tab 5 \tab 3 \tab 3 \tab - \tab 2 \tab - \tab 4 \tab - \tab 2 \tab 9 \tab - \tab 2 \tab 9 \tab 4 \tab 11 \tab - \tab 10 \tab 1 \tab - \tab - \tab 5 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 11 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 7 \tab 6 \tab 16 \tab 6 \tab 2 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 5 \tab - \tab 2 \tab 7 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 6 \tab 1 \tab 1 \tab 10 \tab 2 \tab 6 \tab 12 \tab 4 \tab 1 \tab 16 \tab 7 \tab 4 \tab 7 \tab 8 \tab 1 \tab 38 \tab 1 \tab 2 \tab 1 \tab 14 \tab 11 \tab 9 \tab 7 \tab 11 \tab 3 \tab 5 \tab 17 \tab 14 \tab 26 \tab 2 \tab 3 \tab 1 \tab 4 \tab 3 \tab 2 \tab - \tab 1 \tab 5 \tab - \tab 6 \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 25 \tab 33 \tab 1 \tab 7 \tab 1 \tab 4 \tab 4 \tab 13 \tab 2 \tab 5 \tab 1 \tab 8 \tab 7 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 3 \tab 1 \tab 3 \tab 4 \tab 2 \tab 41 \tab 3 \tab 5 \tab 9 \tab 5 \tab - \tab - \tab - \tab 13 \tab 2 \tab - \tab - \tab 2 \tab 35 \tab 1 \tab 1 \tab 8 \tab 17 \tab - \tab - \tab 3 \tab 21 \tab 237 \tab 45 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 22 \tab 9 \tab 10 \tab 246 \tab 12 \tab 9 \tab 9 \tab 5 \tab 8 \tab 4 \tab 45 \tab - \tab 2 \tab - \tab 4 \tab 15 \tab 1 \tab 20 \tab - \tab - \tab 2 \tab 8 \tab 6 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab 11 \tab - \tab 2 \tab 5 \tab 3 \tab 3 \tab 3 \tab 7 \tab 9 \tab 1 \tab 5 \tab - \tab - \tab 1 \tab 12 \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab 11 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 10 \tab 8 \tab 30 \tab 5 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 6 \tab - \tab - \tab 1 \tab - \tab 10 \tab 2 \tab - \tab 17 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab 6 \tab 8 \tab 1 \tab 10 \tab 5 \tab 8 \tab - \tab 24 \tab 7 \tab 12 \tab 1 \tab 18 \tab 4 \tab 46 \tab - \tab 2 \tab 3 \tab 16 \tab 10 \tab 11 \tab 12 \tab 9 \tab 7 \tab 2 \tab 18 \tab 10 \tab 31 \tab 4 \tab 2 \tab - \tab 6 \tab 6 \tab 2 \tab 2 \tab 4 \tab 3 \tab - \tab 9 \tab 7 \tab - \tab 9 \tab 1 \tab - \tab 3 \tab 30 \tab 34 \tab 3 \tab 5 \tab 5 \tab 4 \tab 4 \tab 28 \tab 1 \tab 4 \tab - \tab 6 \tab 2 \tab - \tab 3 \tab 5 \tab 1 \tab 2 \tab 2 \tab 2 \tab 8 \tab 7 \tab 7 \tab 67 \tab 10 \tab 15 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 27 \tab - \tab 3 \tab 3 \tab 3 \tab 57 \tab 1 \tab 4 \tab 13 \tab 28 \tab - \tab 1 \tab 1 \tab 34 \tab 353 \tab 29 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 6 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 16 \tab 10 \tab 13 \tab 176 \tab 23 \tab 7 \tab 6 \tab 8 \tab 4 \tab 6 \tab 53 \tab 2 \tab 1 \tab 1 \tab 4 \tab 9 \tab 3 \tab 28 \tab 2 \tab 1 \tab 5 \tab 6 \tab 7 \tab 4 \tab 15 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab 8 \tab 5 \tab 1 \tab 11 \tab 4 \tab 10 \tab - \tab 10 \tab - \tab - \tab 1 \tab 12 \tab 1 \tab - \tab - \tab - \tab 1 \tab 7 \tab - \tab 4 \tab - \tab - \tab 6 \tab 8 \tab 3 \tab 15 \tab 4 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 5 \tab 2 \tab 9 \tab 4 \tab - \tab 17 \tab - \tab 6 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 11 \tab - \tab 5 \tab 10 \tab 7 \tab 1 \tab 23 \tab 8 \tab 6 \tab 2 \tab 14 \tab 3 \tab 48 \tab 3 \tab 3 \tab 2 \tab 8 \tab 13 \tab 10 \tab 12 \tab 10 \tab 5 \tab 12 \tab 24 \tab 17 \tab 25 \tab 2 \tab 4 \tab 3 \tab 4 \tab 3 \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab 9 \tab 3 \tab - \tab 6 \tab 2 \tab 1 \tab 2 \tab 44 \tab 24 \tab 4 \tab 4 \tab 1 \tab 1 \tab 8 \tab 18 \tab 2 \tab 8 \tab 1 \tab 6 \tab 5 \tab 1 \tab 2 \tab 4 \tab - \tab 6 \tab 3 \tab 2 \tab 6 \tab 2 \tab 3 \tab 57 \tab 4 \tab 10 \tab 10 \tab 2 \tab 2 \tab 1 \tab - \tab 21 \tab 1 \tab 1 \tab 2 \tab 2 \tab 41 \tab 1 \tab 2 \tab 5 \tab 18 \tab 1 \tab 3 \tab 3 \tab 23 \tab 220 \tab 20 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 4 \tab 2 \tab 2 \tab 1 \tab 9 \tab 5 \tab 2 \tab - \tab 3 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 23 \tab 13 \tab 12 \tab 219 \tab 21 \tab 7 \tab 7 \tab 7 \tab 5 \tab 6 \tab 58 \tab 3 \tab 1 \tab 1 \tab 2 \tab 11 \tab - \tab 21 \tab - \tab 3 \tab 2 \tab 4 \tab 9 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 8 \tab 3 \tab - \tab 3 \tab 1 \tab 2 \tab 5 \tab 3 \tab 14 \tab - \tab 5 \tab - \tab 1 \tab - \tab 6 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 9 \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 11 \tab 6 \tab 21 \tab 4 \tab 2 \tab 3 \tab - \tab 5 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 11 \tab - \tab - \tab 8 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 6 \tab 7 \tab 9 \tab - \tab 4 \tab 7 \tab 3 \tab - \tab 14 \tab 5 \tab 7 \tab - \tab 9 \tab 2 \tab 29 \tab - \tab 1 \tab - \tab 17 \tab 14 \tab 8 \tab 16 \tab 12 \tab 5 \tab 7 \tab 11 \tab 18 \tab 19 \tab 5 \tab 1 \tab 2 \tab 5 \tab 4 \tab 3 \tab 1 \tab 2 \tab 2 \tab - \tab 9 \tab 1 \tab - \tab 3 \tab 4 \tab 3 \tab 1 \tab 31 \tab 22 \tab 3 \tab 2 \tab - \tab - \tab 6 \tab 22 \tab 5 \tab 6 \tab - \tab 6 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 2 \tab 4 \tab 6 \tab 3 \tab 1 \tab 47 \tab 6 \tab 12 \tab 4 \tab 4 \tab 1 \tab 1 \tab - \tab 12 \tab - \tab - \tab 2 \tab 1 \tab 34 \tab 1 \tab 5 \tab 6 \tab 24 \tab 1 \tab - \tab - \tab 22 \tab 218 \tab 3 \tab - \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 6 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 17 \tab 10 \tab 15 \tab 215 \tab 21 \tab 10 \tab 7 \tab 6 \tab 8 \tab 7 \tab 58 \tab 1 \tab 1 \tab 2 \tab - \tab 9 \tab 1 \tab 16 \tab - \tab 4 \tab 2 \tab 7 \tab 4 \tab 3 \tab 8 \tab 5 \tab 4 \tab - \tab 6 \tab - \tab 1 \tab 10 \tab 3 \tab 3 \tab 5 \tab 4 \tab 13 \tab 1 \tab 8 \tab - \tab 1 \tab 1 \tab 13 \tab - \tab - \tab 2 \tab - \tab 1 \tab 6 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 10 \tab 7 \tab 15 \tab 7 \tab 2 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 5 \tab 2 \tab 4 \tab 6 \tab - \tab 15 \tab - \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 5 \tab 5 \tab - \tab 6 \tab 11 \tab 2 \tab - \tab 18 \tab 6 \tab 3 \tab - \tab 11 \tab 5 \tab 41 \tab 2 \tab 4 \tab - \tab 10 \tab 14 \tab 6 \tab 10 \tab 12 \tab 2 \tab 6 \tab 18 \tab 16 \tab 25 \tab 8 \tab 1 \tab 6 \tab 4 \tab 3 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 14 \tab 9 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 40 \tab 16 \tab 7 \tab 5 \tab 2 \tab 2 \tab 8 \tab 22 \tab 3 \tab 5 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 6 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab 8 \tab 2 \tab 48 \tab - \tab 10 \tab 3 \tab 6 \tab 1 \tab 1 \tab 1 \tab 13 \tab 3 \tab - \tab - \tab 3 \tab 37 \tab - \tab 2 \tab 2 \tab 24 \tab - \tab 1 \tab 3 \tab 21 \tab 204 \tab 44 \tab - \tab 2 \tab 3 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab 4 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 14 \tab 9 \tab 15 \tab 192 \tab 16 \tab 8 \tab 9 \tab 7 \tab 5 \tab 6 \tab 52 \tab - \tab 1 \tab 1 \tab 3 \tab 11 \tab 1 \tab 29 \tab - \tab 1 \tab 2 \tab 5 \tab 6 \tab 4 \tab 4 \tab - \tab 2 \tab - \tab 7 \tab 2 \tab - \tab 5 \tab 3 \tab 2 \tab 7 \tab 3 \tab 4 \tab - \tab 7 \tab 1 \tab - \tab 2 \tab 9 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 10 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 8 \tab 6 \tab 27 \tab 5 \tab 3 \tab 1 \tab - \tab 1 \tab 7 \tab 3 \tab - \tab - \tab 2 \tab - \tab 8 \tab 1 \tab - \tab 10 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 5 \tab 4 \tab 2 \tab 5 \tab 13 \tab 6 \tab 1 \tab 29 \tab 8 \tab 2 \tab - \tab 4 \tab 4 \tab 36 \tab 1 \tab - \tab 1 \tab 13 \tab 13 \tab 8 \tab 7 \tab 14 \tab 6 \tab 8 \tab 31 \tab 16 \tab 17 \tab 1 \tab 5 \tab 2 \tab 3 \tab 4 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab 7 \tab 2 \tab - \tab 5 \tab 32 \tab 11 \tab 1 \tab 3 \tab 4 \tab 2 \tab 4 \tab 23 \tab 3 \tab 3 \tab 2 \tab 6 \tab 4 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 2 \tab 2 \tab 7 \tab 4 \tab 5 \tab 39 \tab 4 \tab 6 \tab 5 \tab 9 \tab 1 \tab - \tab 1 \tab 13 \tab - \tab 1 \tab 3 \tab 1 \tab 42 \tab - \tab - \tab 6 \tab 18 \tab - \tab - \tab 3 \tab 18 \tab 246 \tab 7 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 5 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 26 \tab 9 \tab 9 \tab 250 \tab 21 \tab 7 \tab 8 \tab 7 \tab 5 \tab 6 \tab 47 \tab 1 \tab 2 \tab 1 \tab 3 \tab 9 \tab - \tab 21 \tab 1 \tab 1 \tab 5 \tab 7 \tab 10 \tab - \tab 7 \tab 1 \tab 1 \tab - \tab 6 \tab - \tab - \tab 12 \tab 2 \tab 2 \tab 8 \tab 11 \tab 20 \tab 1 \tab 9 \tab 1 \tab - \tab 2 \tab 11 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 7 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 12 \tab 4 \tab 16 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 2 \tab 14 \tab 3 \tab 2 \tab 9 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 6 \tab 8 \tab 1 \tab 2 \tab - \tab 8 \tab 7 \tab 7 \tab - \tab 15 \tab 5 \tab 5 \tab - \tab 10 \tab 4 \tab 50 \tab 4 \tab 1 \tab 1 \tab 12 \tab 10 \tab 13 \tab 12 \tab 20 \tab 3 \tab 6 \tab 15 \tab 18 \tab 22 \tab 1 \tab 6 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 9 \tab 4 \tab - \tab 3 \tab 1 \tab 1 \tab 6 \tab 33 \tab 28 \tab 3 \tab 1 \tab 2 \tab 2 \tab 2 \tab 33 \tab - \tab 3 \tab 1 \tab 6 \tab 2 \tab 2 \tab 1 \tab 4 \tab 4 \tab 3 \tab 1 \tab 2 \tab 5 \tab 7 \tab 5 \tab 69 \tab 3 \tab 8 \tab 7 \tab 4 \tab - \tab 2 \tab - \tab 15 \tab 4 \tab 3 \tab - \tab 2 \tab 65 \tab 1 \tab 2 \tab 5 \tab 34 \tab 1 \tab - \tab 1 \tab 25 \tab 319 \tab 63 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 4 \tab 2 \tab - \tab 3 \tab 2 \tab - \tab 6 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 21 \tab 14 \tab 18 \tab 183 \tab 23 \tab 6 \tab 4 \tab 6 \tab 8 \tab 3 \tab 72 \tab - \tab 2 \tab 5 \tab 3 \tab 11 \tab 1 \tab 24 \tab - \tab 3 \tab 5 \tab 5 \tab 7 \tab 2 \tab 8 \tab 3 \tab - \tab - \tab 5 \tab 1 \tab 2 \tab 8 \tab 2 \tab 2 \tab 7 \tab 6 \tab 14 \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab 10 \tab 5 \tab - \tab - \tab - \tab - \tab 7 \tab - \tab - \tab 2 \tab 1 \tab - \tab 5 \tab 2 \tab 23 \tab 8 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab 8 \tab 2 \tab 1 \tab 7 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 3 \tab 6 \tab 3 \tab 4 \tab - \tab 3 \tab 11 \tab 5 \tab 1 \tab 22 \tab 6 \tab 9 \tab 1 \tab 5 \tab 1 \tab 33 \tab - \tab 2 \tab - \tab 9 \tab 11 \tab 7 \tab 7 \tab 11 \tab 3 \tab 7 \tab 23 \tab 18 \tab 33 \tab 1 \tab 2 \tab 2 \tab 5 \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 6 \tab 4 \tab - \tab 5 \tab 1 \tab - \tab 4 \tab 24 \tab 11 \tab 2 \tab 4 \tab 3 \tab - \tab 6 \tab 28 \tab 3 \tab 6 \tab 1 \tab 3 \tab 3 \tab 2 \tab - \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab 4 \tab 4 \tab 55 \tab 7 \tab 7 \tab 1 \tab 6 \tab 2 \tab 2 \tab - \tab 16 \tab 1 \tab 2 \tab 1 \tab 4 \tab 37 \tab 1 \tab 1 \tab 6 \tab 22 \tab - \tab 1 \tab 2 \tab 18 \tab 226 \tab 1 \tab - \tab 6 \tab - \tab 4 \tab 3 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 7 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 26 \tab 12 \tab 16 \tab 194 \tab 20 \tab 7 \tab 7 \tab 5 \tab 6 \tab 7 \tab 47 \tab - \tab 1 \tab 2 \tab - \tab 9 \tab 3 \tab 23 \tab - \tab 3 \tab 2 \tab 8 \tab 4 \tab 3 \tab 9 \tab 1 \tab 3 \tab - \tab 9 \tab 1 \tab 1 \tab 6 \tab - \tab 4 \tab 11 \tab 9 \tab 8 \tab 1 \tab 6 \tab 1 \tab - \tab 2 \tab 12 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 8 \tab 13 \tab 17 \tab 7 \tab 4 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 8 \tab - \tab - \tab 15 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 6 \tab 4 \tab 6 \tab 1 \tab 8 \tab 10 \tab 1 \tab 1 \tab 17 \tab 6 \tab 5 \tab - \tab 12 \tab 1 \tab 24 \tab 1 \tab 3 \tab 1 \tab 13 \tab 18 \tab 7 \tab 17 \tab 8 \tab 3 \tab 10 \tab 30 \tab 15 \tab 38 \tab 3 \tab 3 \tab 2 \tab 5 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab 5 \tab 3 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 36 \tab 20 \tab 6 \tab 2 \tab 1 \tab - \tab 6 \tab 23 \tab 1 \tab 4 \tab - \tab 5 \tab 6 \tab 2 \tab 5 \tab 5 \tab - \tab 1 \tab 1 \tab 5 \tab 7 \tab 7 \tab 6 \tab 66 \tab 5 \tab 8 \tab 5 \tab 3 \tab 1 \tab - \tab - \tab 7 \tab 2 \tab - \tab 2 \tab 1 \tab 45 \tab 1 \tab - \tab 2 \tab 23 \tab - \tab - \tab 1 \tab 20 \tab 243 \tab 4 \tab 2 \tab 6 \tab 1 \tab 5 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 18 \tab 10 \tab 17 \tab 150 \tab 8 \tab 5 \tab 6 \tab 5 \tab 4 \tab 6 \tab 70 \tab 1 \tab 8 \tab 3 \tab 1 \tab 8 \tab 1 \tab 19 \tab 1 \tab 3 \tab - \tab 6 \tab 7 \tab 5 \tab 9 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 2 \tab 6 \tab 2 \tab 1 \tab 6 \tab 4 \tab 7 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 8 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 7 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 7 \tab 3 \tab 18 \tab 3 \tab 1 \tab - \tab - \tab - \tab 3 \tab 4 \tab - \tab - \tab 8 \tab - \tab 7 \tab - \tab - \tab 12 \tab - \tab 7 \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab 6 \tab 1 \tab 2 \tab 1 \tab 3 \tab 7 \tab 6 \tab 1 \tab 21 \tab 11 \tab 4 \tab 1 \tab 4 \tab - \tab 32 \tab 2 \tab - \tab - \tab 12 \tab 5 \tab 11 \tab 18 \tab 19 \tab 5 \tab 8 \tab 18 \tab 17 \tab 23 \tab 3 \tab 1 \tab 3 \tab 7 \tab 3 \tab - \tab - \tab 4 \tab 3 \tab - \tab 12 \tab 3 \tab - \tab 3 \tab - \tab - \tab 4 \tab 23 \tab 11 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 26 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab 5 \tab 45 \tab 3 \tab 9 \tab 6 \tab 3 \tab 1 \tab 1 \tab - \tab 9 \tab 1 \tab - \tab - \tab 2 \tab 26 \tab - \tab - \tab 2 \tab 23 \tab - \tab - \tab - \tab 23 \tab 233 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 4 \tab 1 \tab 2 \tab 4 \tab 1 \tab - \tab 5 \tab 4 \tab - \tab - \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 27 \tab 5 \tab 9 \tab 173 \tab 20 \tab 9 \tab 5 \tab 13 \tab 8 \tab 5 \tab 68 \tab 2 \tab 3 \tab 3 \tab 4 \tab 12 \tab - \tab 23 \tab - \tab 4 \tab 2 \tab 6 \tab 11 \tab 3 \tab 8 \tab 2 \tab 2 \tab - \tab 7 \tab - \tab 1 \tab 10 \tab 4 \tab 2 \tab 3 \tab 7 \tab 13 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 9 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 6 \tab - \tab - \tab 2 \tab - \tab 1 \tab 6 \tab 5 \tab 29 \tab 5 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 4 \tab - \tab - \tab 4 \tab - \tab 4 \tab 1 \tab 1 \tab 19 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 5 \tab 7 \tab 1 \tab 2 \tab 3 \tab 4 \tab 2 \tab 16 \tab 9 \tab 3 \tab - \tab 7 \tab 3 \tab 33 \tab - \tab 2 \tab - \tab 14 \tab 14 \tab 11 \tab 13 \tab 11 \tab 2 \tab 8 \tab 17 \tab 17 \tab 25 \tab 8 \tab 5 \tab 2 \tab 1 \tab 4 \tab 3 \tab 1 \tab 4 \tab 2 \tab - \tab 4 \tab 3 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 24 \tab 15 \tab 3 \tab 6 \tab 4 \tab 2 \tab 4 \tab 21 \tab 1 \tab 5 \tab - \tab 3 \tab 7 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 4 \tab 2 \tab 2 \tab 54 \tab 4 \tab 10 \tab 6 \tab 3 \tab - \tab 1 \tab 1 \tab 19 \tab 3 \tab 2 \tab 3 \tab - \tab 39 \tab - \tab 1 \tab 2 \tab 24 \tab - \tab - \tab 1 \tab 21 \tab 264 \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 5 \tab - \tab - \tab 2 \tab 2 \tab 10 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 8 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 17 \tab 14 \tab 12 \tab 189 \tab 26 \tab 9 \tab 4 \tab 10 \tab 9 \tab 9 \tab 69 \tab - \tab 4 \tab 2 \tab - \tab 15 \tab - \tab 29 \tab - \tab 4 \tab 5 \tab 10 \tab 6 \tab 2 \tab 7 \tab 1 \tab 2 \tab - \tab 5 \tab 3 \tab 1 \tab 13 \tab 4 \tab 2 \tab 8 \tab 9 \tab 14 \tab - \tab 10 \tab 1 \tab - \tab 1 \tab 10 \tab - \tab - \tab - \tab 1 \tab 5 \tab 13 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 11 \tab 4 \tab 23 \tab 1 \tab 3 \tab 3 \tab - \tab - \tab 3 \tab 5 \tab 1 \tab - \tab 5 \tab 1 \tab 4 \tab 1 \tab 1 \tab 12 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 7 \tab 2 \tab 5 \tab 5 \tab - \tab 9 \tab 15 \tab 4 \tab 3 \tab 26 \tab 9 \tab 9 \tab 1 \tab 7 \tab 2 \tab 35 \tab 2 \tab 3 \tab - \tab 10 \tab 6 \tab 7 \tab 12 \tab 11 \tab 9 \tab 2 \tab 28 \tab 17 \tab 40 \tab 4 \tab 2 \tab 3 \tab 10 \tab 4 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 6 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 30 \tab 4 \tab 1 \tab 3 \tab 3 \tab 2 \tab 6 \tab 25 \tab 3 \tab 5 \tab - \tab 5 \tab 6 \tab - \tab 3 \tab 5 \tab 2 \tab 5 \tab 1 \tab 1 \tab 9 \tab 4 \tab 1 \tab 53 \tab 3 \tab 11 \tab 6 \tab 2 \tab 1 \tab 1 \tab 2 \tab 17 \tab 3 \tab - \tab 1 \tab 1 \tab 41 \tab 1 \tab 2 \tab 5 \tab 12 \tab - \tab 2 \tab 2 \tab 23 \tab 239 \tab 5 \tab 1 \tab 5 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 4 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 5 \tab 3 \tab - \tab 4 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 38 \tab 31 \tab 23 \tab 288 \tab 30 \tab 16 \tab 14 \tab 10 \tab 15 \tab 6 \tab 106 \tab 2 \tab 11 \tab 2 \tab 2 \tab 19 \tab 3 \tab 61 \tab 1 \tab 7 \tab 7 \tab 17 \tab 14 \tab 8 \tab 13 \tab 13 \tab 4 \tab 3 \tab 8 \tab 1 \tab 2 \tab 11 \tab 6 \tab 8 \tab 15 \tab 13 \tab 24 \tab - \tab 23 \tab 1 \tab - \tab 3 \tab 21 \tab 5 \tab 3 \tab 3 \tab 1 \tab 5 \tab 18 \tab - \tab - \tab 5 \tab 1 \tab 3 \tab 19 \tab 7 \tab 41 \tab 12 \tab 4 \tab 5 \tab - \tab 3 \tab 3 \tab 10 \tab 2 \tab 1 \tab 5 \tab - \tab 11 \tab 2 \tab 3 \tab 21 \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 8 \tab 11 \tab 8 \tab 11 \tab 2 \tab 15 \tab 20 \tab 10 \tab 6 \tab 31 \tab 17 \tab 12 \tab - \tab 13 \tab 2 \tab 54 \tab 6 \tab - \tab - \tab 18 \tab 21 \tab 17 \tab 30 \tab 27 \tab 11 \tab 10 \tab 42 \tab 41 \tab 60 \tab 8 \tab 3 \tab 4 \tab 14 \tab 9 \tab 4 \tab 2 \tab 7 \tab 3 \tab 2 \tab 13 \tab 5 \tab - \tab 5 \tab 2 \tab 1 \tab 6 \tab 46 \tab 15 \tab 6 \tab 9 \tab 5 \tab 1 \tab 7 \tab 41 \tab 5 \tab 10 \tab 2 \tab 11 \tab 6 \tab - \tab 2 \tab 2 \tab - \tab 5 \tab 2 \tab 6 \tab 5 \tab 13 \tab 6 \tab 116 \tab 4 \tab 16 \tab 11 \tab 8 \tab - \tab 8 \tab - \tab 44 \tab 7 \tab 1 \tab 1 \tab 1 \tab 58 \tab - \tab 5 \tab 7 \tab 50 \tab 1 \tab - \tab 2 \tab 49 \tab 549 \tab 2 \tab 2 \tab 9 \tab 5 \tab 2 \tab 4 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 5 \tab - \tab 3 \tab 9 \tab 2 \tab 1 \tab 4 \tab 8 \tab 2 \tab 3 \tab 6 \tab - \tab 8 \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab 9 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 29 \tab 30 \tab 30 \tab 235 \tab 43 \tab 16 \tab 8 \tab 8 \tab 12 \tab 14 \tab 110 \tab 5 \tab 11 \tab 7 \tab 2 \tab 20 \tab 2 \tab 57 \tab 2 \tab 5 \tab 5 \tab 14 \tab 24 \tab 8 \tab 12 \tab 7 \tab 5 \tab - \tab 13 \tab 1 \tab 1 \tab 15 \tab 4 \tab 6 \tab 10 \tab 23 \tab 27 \tab 3 \tab 16 \tab 2 \tab 1 \tab 2 \tab 23 \tab 5 \tab 1 \tab 4 \tab 2 \tab 12 \tab 14 \tab 1 \tab 3 \tab - \tab 3 \tab 1 \tab 14 \tab 10 \tab 42 \tab 8 \tab 6 \tab - \tab - \tab 2 \tab 5 \tab 6 \tab 1 \tab 1 \tab 7 \tab - \tab 10 \tab 3 \tab 3 \tab 19 \tab - \tab 11 \tab 1 \tab 1 \tab - \tab 3 \tab 10 \tab 6 \tab 4 \tab 12 \tab 4 \tab 9 \tab 18 \tab 11 \tab 3 \tab 42 \tab 13 \tab 12 \tab 1 \tab 10 \tab 5 \tab 44 \tab 4 \tab 6 \tab 3 \tab 20 \tab 32 \tab 12 \tab 24 \tab 27 \tab 10 \tab 9 \tab 38 \tab 39 \tab 61 \tab 3 \tab 2 \tab 3 \tab 11 \tab 5 \tab 4 \tab - \tab 7 \tab 4 \tab 1 \tab 11 \tab 7 \tab - \tab 4 \tab 3 \tab 2 \tab - \tab 58 \tab 14 \tab 3 \tab 6 \tab 3 \tab 2 \tab 5 \tab 49 \tab 5 \tab 3 \tab 1 \tab 7 \tab 4 \tab 2 \tab 3 \tab 7 \tab 1 \tab 5 \tab 2 \tab 5 \tab 5 \tab 6 \tab 4 \tab 100 \tab 4 \tab 17 \tab 8 \tab 11 \tab 1 \tab 2 \tab - \tab 29 \tab 5 \tab 1 \tab 1 \tab 3 \tab 55 \tab 1 \tab 6 \tab 7 \tab 57 \tab 3 \tab 1 \tab 5 \tab 36 \tab 566 \tab 2 \tab - \tab 13 \tab 3 \tab 6 \tab 2 \tab 2 \tab - \tab 1 \tab 9 \tab 3 \tab 2 \tab 3 \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 5 \tab 6 \tab - \tab 9 \tab 2 \tab - \tab 5 \tab 5 \tab 3 \tab - \tab 2 \tab 3 \tab - \tab 3 \tab 3 \tab - \tab 5 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 8 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 5 \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 41 \tab 18 \tab 27 \tab 227 \tab 37 \tab 25 \tab 5 \tab 14 \tab 13 \tab 14 \tab 103 \tab - \tab 3 \tab 3 \tab 3 \tab 28 \tab 2 \tab 53 \tab 1 \tab 5 \tab 6 \tab 16 \tab 19 \tab 2 \tab 17 \tab 1 \tab 3 \tab 1 \tab 10 \tab 4 \tab 3 \tab 17 \tab 5 \tab 6 \tab 22 \tab 21 \tab 26 \tab - \tab 21 \tab 2 \tab 1 \tab 3 \tab 18 \tab 3 \tab 4 \tab 2 \tab 1 \tab 6 \tab 19 \tab 1 \tab - \tab 3 \tab 1 \tab 3 \tab 21 \tab 12 \tab 29 \tab 6 \tab 1 \tab 9 \tab - \tab 3 \tab 4 \tab 5 \tab 1 \tab - \tab 6 \tab 1 \tab 12 \tab 2 \tab 2 \tab 25 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 10 \tab 13 \tab 4 \tab 4 \tab 3 \tab 10 \tab 23 \tab 13 \tab 4 \tab 37 \tab 16 \tab 10 \tab 2 \tab 21 \tab 4 \tab 53 \tab 2 \tab 4 \tab 2 \tab 17 \tab 29 \tab 12 \tab 32 \tab 15 \tab 9 \tab 14 \tab 48 \tab 43 \tab 53 \tab 6 \tab 5 \tab 6 \tab 15 \tab 6 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 15 \tab 4 \tab - \tab 2 \tab 2 \tab 3 \tab 7 \tab 56 \tab 22 \tab 4 \tab 6 \tab 1 \tab 2 \tab 10 \tab 49 \tab 3 \tab 4 \tab 1 \tab 4 \tab 8 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 3 \tab 8 \tab 9 \tab 4 \tab 3 \tab 84 \tab 9 \tab 16 \tab 3 \tab 4 \tab 2 \tab - \tab - \tab 31 \tab 3 \tab 4 \tab - \tab 3 \tab 52 \tab 1 \tab 5 \tab 6 \tab 37 \tab 2 \tab - \tab 1 \tab 25 \tab 568 \tab 1 \tab 2 \tab 11 \tab 4 \tab 5 \tab 3 \tab 2 \tab - \tab 1 \tab 8 \tab 7 \tab 2 \tab 2 \tab 4 \tab 1 \tab - \tab 4 \tab 1 \tab 3 \tab 3 \tab - \tab 5 \tab 5 \tab 1 \tab 4 \tab 6 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 5 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 13 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 34 \tab 21 \tab 18 \tab 245 \tab 30 \tab 12 \tab 15 \tab 11 \tab 11 \tab 14 \tab 91 \tab - \tab 9 \tab 1 \tab 3 \tab 39 \tab 1 \tab 50 \tab 1 \tab 4 \tab 6 \tab 13 \tab 19 \tab 8 \tab 13 \tab 4 \tab 3 \tab - \tab 10 \tab 3 \tab 1 \tab 13 \tab 4 \tab 4 \tab 16 \tab 16 \tab 25 \tab 2 \tab 13 \tab - \tab - \tab 3 \tab 20 \tab 5 \tab - \tab - \tab - \tab 5 \tab 15 \tab 1 \tab - \tab 1 \tab 1 \tab 7 \tab 15 \tab 9 \tab 46 \tab 7 \tab 2 \tab 3 \tab - \tab 5 \tab 5 \tab 7 \tab 1 \tab 3 \tab 3 \tab - \tab 15 \tab 1 \tab 1 \tab 14 \tab - \tab 12 \tab 1 \tab 2 \tab - \tab - \tab 8 \tab 11 \tab 8 \tab 7 \tab 1 \tab 15 \tab 16 \tab 13 \tab 3 \tab 28 \tab 15 \tab 10 \tab 1 \tab 10 \tab 2 \tab 46 \tab 2 \tab 10 \tab 4 \tab 22 \tab 18 \tab 8 \tab 23 \tab 20 \tab 11 \tab 11 \tab 29 \tab 34 \tab 61 \tab 3 \tab 9 \tab 3 \tab 12 \tab 4 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 11 \tab 6 \tab 1 \tab 7 \tab 3 \tab 2 \tab 2 \tab 50 \tab 19 \tab 7 \tab 3 \tab 6 \tab 3 \tab 10 \tab 51 \tab 2 \tab 2 \tab 2 \tab 4 \tab 11 \tab 1 \tab 5 \tab 4 \tab 1 \tab 1 \tab 3 \tab 5 \tab 11 \tab 6 \tab 2 \tab 128 \tab 5 \tab 21 \tab 6 \tab 9 \tab 2 \tab 3 \tab 2 \tab 33 \tab 2 \tab 4 \tab 4 \tab 2 \tab 58 \tab - \tab 6 \tab 6 \tab 42 \tab 1 \tab 1 \tab 2 \tab 38 \tab 576 \tab - \tab - \tab 9 \tab 1 \tab 3 \tab 2 \tab 2 \tab - \tab 2 \tab 7 \tab 6 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 5 \tab 5 \tab - \tab 4 \tab 8 \tab 2 \tab 1 \tab 4 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 9 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 11 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab 5 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 54 \tab 17 \tab 24 \tab 200 \tab 34 \tab 19 \tab 4 \tab 14 \tab 14 \tab 13 \tab 100 \tab 2 \tab 8 \tab 6 \tab 4 \tab 25 \tab 3 \tab 49 \tab 2 \tab 5 \tab 5 \tab 15 \tab 12 \tab 5 \tab 10 \tab 7 \tab - \tab 1 \tab 12 \tab 1 \tab 3 \tab 13 \tab 5 \tab 4 \tab 16 \tab 18 \tab 28 \tab 2 \tab 11 \tab 2 \tab - \tab - \tab 16 \tab 3 \tab 1 \tab 2 \tab 3 \tab 2 \tab 15 \tab - \tab 2 \tab 2 \tab 3 \tab 4 \tab 9 \tab 9 \tab 35 \tab 5 \tab 3 \tab 5 \tab - \tab 6 \tab 1 \tab 7 \tab 2 \tab - \tab 5 \tab - \tab 7 \tab 1 \tab 3 \tab 24 \tab - \tab 6 \tab - \tab - \tab - \tab - \tab 5 \tab 15 \tab 8 \tab 5 \tab 1 \tab 9 \tab 28 \tab 12 \tab 2 \tab 29 \tab 13 \tab 13 \tab 1 \tab 10 \tab - \tab 52 \tab - \tab 6 \tab 2 \tab 16 \tab 30 \tab 8 \tab 20 \tab 19 \tab 4 \tab 19 \tab 27 \tab 39 \tab 66 \tab 5 \tab 3 \tab 4 \tab 9 \tab 12 \tab 1 \tab 1 \tab 4 \tab 3 \tab - \tab 11 \tab 5 \tab - \tab 4 \tab 1 \tab 1 \tab 6 \tab 50 \tab 16 \tab 2 \tab 5 \tab 2 \tab 5 \tab 8 \tab 65 \tab 1 \tab 7 \tab 4 \tab 2 \tab 7 \tab 1 \tab 2 \tab 3 \tab 1 \tab 7 \tab 2 \tab 5 \tab 10 \tab 6 \tab 4 \tab 107 \tab 4 \tab 13 \tab 7 \tab 4 \tab 2 \tab 3 \tab 1 \tab 25 \tab 2 \tab 6 \tab 2 \tab 4 \tab 54 \tab - \tab 2 \tab 4 \tab 48 \tab 1 \tab - \tab 1 \tab 36 \tab 589 \tab - \tab 1 \tab 9 \tab 2 \tab 3 \tab 2 \tab 2 \tab 2 \tab 3 \tab 11 \tab 1 \tab 2 \tab 5 \tab 4 \tab 2 \tab - \tab 7 \tab 1 \tab 2 \tab 3 \tab - \tab 3 \tab 6 \tab 3 \tab - \tab 3 \tab 5 \tab 1 \tab 4 \tab 3 \tab 4 \tab 2 \tab - \tab - \tab 5 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 9 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 6 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 52 \tab 21 \tab 24 \tab 171 \tab 24 \tab 11 \tab 6 \tab 8 \tab 12 \tab 14 \tab 88 \tab 4 \tab 7 \tab 3 \tab 4 \tab 28 \tab 2 \tab 51 \tab - \tab 7 \tab 10 \tab 14 \tab 12 \tab 5 \tab 19 \tab 4 \tab 1 \tab 1 \tab 10 \tab 1 \tab 2 \tab 20 \tab 2 \tab 1 \tab 23 \tab 24 \tab 31 \tab 2 \tab 17 \tab 3 \tab - \tab 1 \tab 14 \tab 3 \tab - \tab 2 \tab 1 \tab 5 \tab 15 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 17 \tab 6 \tab 31 \tab 8 \tab 1 \tab 3 \tab - \tab 2 \tab 6 \tab 7 \tab - \tab - \tab 6 \tab - \tab 13 \tab 6 \tab 2 \tab 24 \tab - \tab 9 \tab - \tab - \tab - \tab - \tab 3 \tab 12 \tab 2 \tab 5 \tab 4 \tab 9 \tab 20 \tab 11 \tab 1 \tab 48 \tab 14 \tab 14 \tab 1 \tab 11 \tab 2 \tab 55 \tab 3 \tab 5 \tab 1 \tab 20 \tab 21 \tab 19 \tab 26 \tab 20 \tab 8 \tab 11 \tab 35 \tab 39 \tab 57 \tab 6 \tab 4 \tab 2 \tab 9 \tab 7 \tab 4 \tab - \tab 4 \tab 4 \tab 1 \tab 15 \tab 6 \tab - \tab 4 \tab 1 \tab 3 \tab 3 \tab 57 \tab 25 \tab 6 \tab 6 \tab 5 \tab 4 \tab 16 \tab 50 \tab 2 \tab 8 \tab - \tab 5 \tab 7 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 7 \tab 2 \tab 5 \tab 102 \tab 10 \tab 15 \tab 4 \tab 8 \tab 1 \tab 3 \tab - \tab 30 \tab 4 \tab 4 \tab 1 \tab 5 \tab 62 \tab 1 \tab 6 \tab 2 \tab 62 \tab 1 \tab - \tab - \tab 24 \tab 603 \tab - \tab 2 \tab 15 \tab 7 \tab 7 \tab - \tab 1 \tab 1 \tab - \tab 8 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 1 \tab 4 \tab 4 \tab 1 \tab 5 \tab 9 \tab 1 \tab 1 \tab 3 \tab 3 \tab - \tab 4 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 10 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 6 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 6 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 61 \tab 22 \tab 26 \tab 192 \tab 48 \tab 18 \tab 9 \tab 18 \tab 21 \tab 16 \tab 89 \tab - \tab 7 \tab 1 \tab 2 \tab 25 \tab 1 \tab 52 \tab 2 \tab 7 \tab 8 \tab 15 \tab 16 \tab 5 \tab 16 \tab 9 \tab - \tab 2 \tab 12 \tab 2 \tab 2 \tab 20 \tab 5 \tab 8 \tab 12 \tab 10 \tab 31 \tab 3 \tab 20 \tab 2 \tab 1 \tab 2 \tab 23 \tab 5 \tab 3 \tab - \tab 2 \tab 2 \tab 14 \tab - \tab 4 \tab 3 \tab 5 \tab 3 \tab 23 \tab 7 \tab 33 \tab 9 \tab 2 \tab 7 \tab - \tab 5 \tab 5 \tab 2 \tab - \tab - \tab 3 \tab - \tab 13 \tab 2 \tab 2 \tab 25 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 11 \tab 10 \tab 12 \tab 7 \tab 1 \tab 13 \tab 18 \tab 10 \tab 3 \tab 40 \tab 15 \tab 22 \tab - \tab 15 \tab 3 \tab 53 \tab 5 \tab 4 \tab 4 \tab 24 \tab 31 \tab 15 \tab 32 \tab 25 \tab 10 \tab 12 \tab 25 \tab 38 \tab 67 \tab 2 \tab 12 \tab 9 \tab 19 \tab 10 \tab 1 \tab - \tab 4 \tab 4 \tab - \tab 15 \tab 2 \tab 1 \tab 5 \tab 1 \tab 4 \tab 4 \tab 40 \tab 13 \tab 4 \tab 12 \tab 3 \tab 5 \tab 10 \tab 53 \tab 1 \tab 5 \tab 2 \tab 7 \tab 8 \tab 1 \tab 3 \tab 5 \tab 1 \tab 4 \tab 6 \tab 3 \tab 6 \tab 4 \tab 2 \tab 106 \tab 8 \tab 18 \tab 7 \tab 9 \tab 2 \tab 1 \tab 5 \tab 33 \tab 1 \tab 1 \tab 3 \tab 7 \tab 78 \tab - \tab 7 \tab 2 \tab 60 \tab - \tab 1 \tab 2 \tab 34 \tab 2107 \tab - \tab 3 \tab 23 \tab 1 \tab 8 \tab 3 \tab 2 \tab 1 \tab 2 \tab 8 \tab 2 \tab 1 \tab 5 \tab 3 \tab 3 \tab - \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 6 \tab 2 \tab - \tab 6 \tab 2 \tab 2 \tab 1 \tab 6 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab 2 \tab - \tab 10 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 2 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 49 \tab 16 \tab 16 \tab 127 \tab 26 \tab 12 \tab 14 \tab 8 \tab 8 \tab 6 \tab 45 \tab - \tab 7 \tab 7 \tab 5 \tab 23 \tab 1 \tab 46 \tab 1 \tab 3 \tab 5 \tab 7 \tab 17 \tab 5 \tab 16 \tab 1 \tab 4 \tab 2 \tab 7 \tab - \tab 3 \tab 10 \tab 2 \tab 3 \tab 9 \tab 8 \tab 19 \tab 5 \tab 7 \tab 1 \tab 3 \tab 3 \tab 9 \tab 3 \tab 1 \tab 2 \tab - \tab 3 \tab 11 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 24 \tab 11 \tab 20 \tab 8 \tab 2 \tab 2 \tab - \tab 2 \tab 4 \tab 5 \tab 2 \tab 1 \tab 6 \tab - \tab 12 \tab - \tab 1 \tab 8 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 6 \tab 9 \tab 6 \tab 5 \tab 4 \tab 11 \tab 15 \tab 7 \tab 2 \tab 25 \tab 14 \tab 9 \tab 2 \tab 4 \tab 2 \tab 23 \tab 3 \tab - \tab 3 \tab 18 \tab 28 \tab 7 \tab 8 \tab 15 \tab 6 \tab 7 \tab 40 \tab 26 \tab 47 \tab 3 \tab - \tab 4 \tab 9 \tab 7 \tab - \tab - \tab 4 \tab 2 \tab - \tab 10 \tab 6 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 32 \tab 14 \tab 4 \tab 7 \tab 2 \tab 1 \tab 6 \tab 44 \tab 2 \tab 6 \tab 1 \tab 4 \tab 7 \tab - \tab 2 \tab 5 \tab 1 \tab 4 \tab 1 \tab 3 \tab 3 \tab 3 \tab 3 \tab 69 \tab 4 \tab 10 \tab 4 \tab 7 \tab 2 \tab 3 \tab - \tab 24 \tab 3 \tab 2 \tab 1 \tab 4 \tab 33 \tab 1 \tab 9 \tab 1 \tab 42 \tab 2 \tab - \tab - \tab 21 \tab 398 \tab - \tab 1 \tab 6 \tab 4 \tab 6 \tab 2 \tab 3 \tab - \tab 1 \tab 5 \tab 6 \tab 2 \tab 3 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 3 \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab 5 \tab 1 \tab 5 \tab - \tab - \tab 6 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab 40 \tab 19 \tab 21 \tab 113 \tab 26 \tab 14 \tab 9 \tab 1 \tab 6 \tab 14 \tab 73 \tab 3 \tab 4 \tab - \tab 2 \tab 16 \tab - \tab 50 \tab 1 \tab 2 \tab 2 \tab 11 \tab 14 \tab 7 \tab 8 \tab 2 \tab 1 \tab - \tab 6 \tab 4 \tab 1 \tab 20 \tab 7 \tab 3 \tab 19 \tab 11 \tab 21 \tab 3 \tab 10 \tab 1 \tab 1 \tab 3 \tab 6 \tab 3 \tab - \tab 2 \tab - \tab 4 \tab 18 \tab - \tab 1 \tab 2 \tab 2 \tab 4 \tab 10 \tab 6 \tab 25 \tab 10 \tab 3 \tab 4 \tab - \tab 4 \tab 8 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 8 \tab - \tab 2 \tab 10 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab 12 \tab 4 \tab 7 \tab - \tab 4 \tab 13 \tab 7 \tab 1 \tab 24 \tab 14 \tab 7 \tab - \tab 6 \tab 4 \tab 31 \tab 2 \tab 2 \tab - \tab 18 \tab 18 \tab 8 \tab 19 \tab 13 \tab 5 \tab 3 \tab 21 \tab 29 \tab 56 \tab 1 \tab 5 \tab 3 \tab 11 \tab 6 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 6 \tab 2 \tab - \tab 5 \tab 2 \tab 2 \tab 1 \tab 33 \tab 10 \tab 5 \tab 6 \tab 3 \tab 6 \tab 12 \tab 25 \tab - \tab 4 \tab 1 \tab 7 \tab 3 \tab - \tab 2 \tab 5 \tab - \tab 3 \tab 3 \tab 4 \tab 10 \tab 2 \tab - \tab 85 \tab 6 \tab 17 \tab 9 \tab 7 \tab 2 \tab 2 \tab - \tab 26 \tab 6 \tab 4 \tab 2 \tab 4 \tab 39 \tab - \tab 8 \tab 2 \tab 43 \tab - \tab - \tab 1 \tab 26 \tab 402 \tab - \tab 1 \tab 8 \tab 2 \tab 7 \tab 1 \tab - \tab 1 \tab - \tab 7 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab 6 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 16 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 75 \tab 15 \tab 13 \tab 113 \tab 20 \tab 10 \tab 3 \tab 5 \tab 7 \tab 9 \tab 52 \tab 2 \tab 5 \tab 5 \tab 3 \tab 23 \tab 1 \tab 52 \tab 2 \tab 2 \tab 5 \tab 10 \tab 11 \tab 4 \tab 11 \tab - \tab 4 \tab 2 \tab 5 \tab 4 \tab 3 \tab 22 \tab 3 \tab 2 \tab 21 \tab 15 \tab 10 \tab 1 \tab 8 \tab - \tab 2 \tab 1 \tab 16 \tab 2 \tab - \tab 2 \tab 2 \tab 4 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 15 \tab 2 \tab 27 \tab 8 \tab 4 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 12 \tab 3 \tab - \tab 13 \tab - \tab 7 \tab - \tab 2 \tab - \tab - \tab 4 \tab 10 \tab 5 \tab 7 \tab 3 \tab 12 \tab 14 \tab 10 \tab 3 \tab 26 \tab 10 \tab 6 \tab - \tab 9 \tab 2 \tab 30 \tab 1 \tab 4 \tab - \tab 12 \tab 16 \tab 11 \tab 15 \tab 17 \tab 6 \tab 7 \tab 25 \tab 19 \tab 43 \tab 3 \tab 6 \tab 2 \tab 9 \tab 4 \tab - \tab - \tab 5 \tab - \tab - \tab 7 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 22 \tab 10 \tab 2 \tab 2 \tab 1 \tab 1 \tab 7 \tab 55 \tab - \tab 4 \tab - \tab 4 \tab 5 \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab 2 \tab 4 \tab 6 \tab - \tab 60 \tab 4 \tab 8 \tab 7 \tab 7 \tab - \tab 5 \tab 2 \tab 31 \tab 4 \tab 2 \tab 1 \tab 2 \tab 51 \tab - \tab 1 \tab 1 \tab 58 \tab - \tab 1 \tab - \tab 28 \tab 385 \tab - \tab 2 \tab 7 \tab 2 \tab 5 \tab 2 \tab 2 \tab - \tab 1 \tab 8 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 4 \tab 5 \tab - \tab 5 \tab 3 \tab 1 \tab - \tab - \tab 6 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 11 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 55 \tab 24 \tab 39 \tab 131 \tab 27 \tab 22 \tab 11 \tab 10 \tab 7 \tab 13 \tab 79 \tab 2 \tab 14 \tab 4 \tab 10 \tab 29 \tab 3 \tab 39 \tab 2 \tab 1 \tab 6 \tab 14 \tab 13 \tab 3 \tab 17 \tab 5 \tab 4 \tab 2 \tab 11 \tab 1 \tab 3 \tab 20 \tab 4 \tab 1 \tab 19 \tab 14 \tab 27 \tab 5 \tab 17 \tab 2 \tab - \tab 5 \tab 16 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 14 \tab 1 \tab 1 \tab 4 \tab - \tab 2 \tab 12 \tab 3 \tab 47 \tab 11 \tab 3 \tab 3 \tab - \tab 6 \tab 4 \tab 6 \tab - \tab - \tab 3 \tab - \tab 8 \tab 5 \tab 3 \tab 14 \tab - \tab 10 \tab 1 \tab - \tab - \tab - \tab 1 \tab 8 \tab 5 \tab 5 \tab 1 \tab 8 \tab 22 \tab 12 \tab 3 \tab 39 \tab 18 \tab 6 \tab - \tab 11 \tab 1 \tab 42 \tab 6 \tab 3 \tab 1 \tab 28 \tab 16 \tab 16 \tab 20 \tab 21 \tab 5 \tab 12 \tab 30 \tab 22 \tab 61 \tab 3 \tab 4 \tab 6 \tab 9 \tab 3 \tab 1 \tab - \tab 3 \tab 5 \tab - \tab 13 \tab 4 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 35 \tab 15 \tab 3 \tab 4 \tab 4 \tab 1 \tab 3 \tab 54 \tab - \tab 2 \tab - \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab 7 \tab 2 \tab 1 \tab 90 \tab 5 \tab 16 \tab 9 \tab 8 \tab 1 \tab 1 \tab - \tab 30 \tab 5 \tab 1 \tab 2 \tab 3 \tab 57 \tab - \tab 2 \tab 4 \tab 56 \tab 1 \tab 1 \tab 3 \tab 33 \tab 524 \tab - \tab - \tab 4 \tab 2 \tab 5 \tab 4 \tab 1 \tab 2 \tab 2 \tab 8 \tab 1 \tab 2 \tab 5 \tab 2 \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab 8 \tab 5 \tab - \tab - \tab 4 \tab 2 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 7 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 16 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 41 \tab 18 \tab 34 \tab 117 \tab 17 \tab 20 \tab 11 \tab 10 \tab 9 \tab 20 \tab 87 \tab 3 \tab 3 \tab 3 \tab 2 \tab 25 \tab 3 \tab 62 \tab 4 \tab 2 \tab 2 \tab 15 \tab 11 \tab 5 \tab 29 \tab 6 \tab 1 \tab 1 \tab 21 \tab 8 \tab 6 \tab 17 \tab 3 \tab 5 \tab 16 \tab 21 \tab 28 \tab 7 \tab 15 \tab 1 \tab 1 \tab 4 \tab 12 \tab 6 \tab - \tab 2 \tab 4 \tab 4 \tab 16 \tab 1 \tab - \tab 2 \tab 5 \tab 5 \tab 11 \tab 6 \tab 26 \tab 8 \tab 3 \tab 3 \tab - \tab 4 \tab 5 \tab 4 \tab 1 \tab - \tab 2 \tab 1 \tab 8 \tab 3 \tab 4 \tab 16 \tab 1 \tab 9 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 17 \tab 5 \tab 9 \tab 1 \tab 9 \tab 29 \tab 12 \tab 9 \tab 35 \tab 15 \tab 8 \tab 2 \tab 4 \tab 5 \tab 38 \tab 5 \tab 3 \tab 2 \tab 24 \tab 25 \tab 20 \tab 19 \tab 27 \tab 5 \tab 20 \tab 34 \tab 30 \tab 72 \tab 7 \tab 5 \tab 2 \tab 15 \tab 8 \tab 2 \tab 2 \tab 2 \tab 5 \tab - \tab 7 \tab 8 \tab 1 \tab 4 \tab 1 \tab 2 \tab 3 \tab 36 \tab 11 \tab 3 \tab 2 \tab 2 \tab 5 \tab 6 \tab 58 \tab 1 \tab 3 \tab 1 \tab 2 \tab 10 \tab - \tab 3 \tab 4 \tab - \tab - \tab 4 \tab 5 \tab 14 \tab 9 \tab 2 \tab 103 \tab 9 \tab 15 \tab 3 \tab 7 \tab 2 \tab 2 \tab - \tab 53 \tab 6 \tab 5 \tab 5 \tab 4 \tab 77 \tab - \tab 2 \tab 3 \tab 76 \tab - \tab 3 \tab 1 \tab 31 \tab 625 \tab - \tab 2 \tab 11 \tab 4 \tab 8 \tab 5 \tab 3 \tab - \tab 6 \tab 11 \tab 3 \tab 2 \tab 1 \tab 2 \tab 7 \tab 2 \tab 5 \tab - \tab - \tab 2 \tab - \tab 5 \tab 7 \tab 1 \tab 6 \tab 9 \tab - \tab - \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 7 \tab 3 \tab 2 \tab - \tab - \tab - \tab 25 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab 3 \tab - \tab 4 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2867 \cr
   2018 \tab 77 \tab 28 \tab 23 \tab 98 \tab 48 \tab 9 \tab 11 \tab 19 \tab 11 \tab 11 \tab 48 \tab 1 \tab 10 \tab 1 \tab 6 \tab 21 \tab 1 \tab 29 \tab 2 \tab - \tab 7 \tab 10 \tab 5 \tab 4 \tab 7 \tab 2 \tab 5 \tab 3 \tab 21 \tab 3 \tab 2 \tab 19 \tab 6 \tab 3 \tab 13 \tab 12 \tab 21 \tab 8 \tab 8 \tab 1 \tab 2 \tab - \tab 10 \tab 2 \tab 1 \tab 2 \tab 2 \tab 3 \tab 12 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 17 \tab 10 \tab 24 \tab 6 \tab 4 \tab 4 \tab - \tab 5 \tab 7 \tab 4 \tab 1 \tab - \tab 1 \tab 2 \tab 11 \tab 3 \tab 2 \tab 20 \tab - \tab 10 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 15 \tab 4 \tab 4 \tab 3 \tab 16 \tab 20 \tab 5 \tab 7 \tab 33 \tab 19 \tab 8 \tab - \tab 5 \tab 5 \tab 34 \tab 3 \tab 6 \tab - \tab 11 \tab 11 \tab 7 \tab 15 \tab 23 \tab 6 \tab 8 \tab 21 \tab 28 \tab 56 \tab - \tab 6 \tab 2 \tab 5 \tab 10 \tab 2 \tab 1 \tab 4 \tab 3 \tab - \tab 6 \tab 3 \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab 26 \tab 9 \tab 6 \tab 2 \tab 1 \tab 8 \tab 9 \tab 37 \tab 1 \tab 2 \tab 3 \tab 3 \tab 4 \tab - \tab 2 \tab 4 \tab - \tab 2 \tab 3 \tab 4 \tab 12 \tab 2 \tab 2 \tab 94 \tab 7 \tab 13 \tab 6 \tab 5 \tab 3 \tab 1 \tab 2 \tab 33 \tab - \tab 3 \tab 1 \tab 5 \tab 50 \tab 1 \tab 1 \tab 1 \tab 42 \tab 3 \tab 3 \tab 1 \tab 34 \tab 498 \tab - \tab 5 \tab 5 \tab 4 \tab 5 \tab 2 \tab 2 \tab 2 \tab 3 \tab 6 \tab - \tab 3 \tab 1 \tab 3 \tab 5 \tab - \tab 10 \tab 2 \tab 1 \tab 4 \tab - \tab 5 \tab 11 \tab - \tab 2 \tab 4 \tab - \tab - \tab 3 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 10 \tab 1 \tab - \tab - \tab 1 \tab - \tab 15 \tab 7 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 4 \tab 1 \tab - \tab 2 \tab 6 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 1 \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 1003 \tab 437 \tab 529 \tab 6388 \tab 755 \tab 343 \tab 236 \tab 259 \tab 238 \tab 275 \tab 2035 \tab 43 \tab 141 \tab 75 \tab 94 \tab 480 \tab 48 \tab 1020 \tab 26 \tab 92 \tab 122 \tab 268 \tab 301 \tab 112 \tab 307 \tab 88 \tab 65 \tab 23 \tab 248 \tab 46 \tab 49 \tab 355 \tab 102 \tab 103 \tab 313 \tab 298 \tab 521 \tab 49 \tab 296 \tab 30 \tab 20 \tab 54 \tab 382 \tab 92 \tab 20 \tab 41 \tab 44 \tab 78 \tab 352 \tab 12 \tab 29 \tab 48 \tab 54 \tab 68 \tab 368 \tab 202 \tab 785 \tab 195 \tab 70 \tab 71 \tab 4 \tab 76 \tab 98 \tab 116 \tab 22 \tab 15 \tab 127 \tab 14 \tab 280 \tab 60 \tab 42 \tab 456 \tab 5 \tab 144 \tab 15 \tab 12 \tab 8 \tab 16 \tab 148 \tab 227 \tab 153 \tab 209 \tab 42 \tab 240 \tab 444 \tab 223 \tab 65 \tab 781 \tab 320 \tab 276 \tab 42 \tab 296 \tab 76 \tab 1290 \tab 69 \tab 81 \tab 47 \tab 471 \tab 493 \tab 349 \tab 498 \tab 482 \tab 177 \tab 262 \tab 811 \tab 692 \tab 1257 \tab 114 \tab 114 \tab 84 \tab 233 \tab 148 \tab 55 \tab 19 \tab 113 \tab 78 \tab 11 \tab 265 \tab 135 \tab 10 \tab 125 \tab 47 \tab 45 \tab 79 \tab 1110 \tab 648 \tab 112 \tab 143 \tab 88 \tab 72 \tab 240 \tab 1041 \tab 57 \tab 156 \tab 33 \tab 172 \tab 173 \tab 29 \tab 74 \tab 107 \tab 29 \tab 82 \tab 61 \tab 92 \tab 216 \tab 156 \tab 101 \tab 2197 \tab 147 \tab 368 \tab 184 \tab 156 \tab 38 \tab 69 \tab 27 \tab 693 \tab 71 \tab 56 \tab 47 \tab 79 \tab 1572 \tab 17 \tab 91 \tab 173 \tab 1088 \tab 22 \tab 22 \tab 50 \tab 849 \tab 20476 \tab 810 \tab 34 \tab 190 \tab 58 \tab 101 \tab 57 \tab 46 \tab 15 \tab 37 \tab 136 \tab 50 \tab 29 \tab 46 \tab 45 \tab 42 \tab 10 \tab 92 \tab 22 \tab 42 \tab 46 \tab 6 \tab 85 \tab 82 \tab 17 \tab 90 \tab 108 \tab 47 \tab 16 \tab 107 \tab 75 \tab 26 \tab 38 \tab 42 \tab 13 \tab 157 \tab 43 \tab 20 \tab 46 \tab 16 \tab 8 \tab 184 \tab 41 \tab 14 \tab 8 \tab 2 \tab 20 \tab 13 \tab 30 \tab 8 \tab 20 \tab 29 \tab 77 \tab 15 \tab 30 \tab 24 \tab 6 \tab 11 \tab 18 \tab 54 \tab 11 \tab 14 \tab 2 \tab 9 \tab 7 \tab 7 \tab 3 \tab 9 \tab 13 \tab 24 \tab 11 \tab 10 \tab 25 \tab 10 \tab 6 \tab 13 \tab 5 \tab 44 \tab 12 \tab 15 \tab 13 \tab 11 \tab 10 \tab 3 \tab 14 \tab 25 \tab 3 \tab 3 \tab 15 \tab 9 \tab 4 \tab 19 \tab 2 \tab 7 \tab 2 \tab 2 \tab 4 \tab 3 \tab 3 \tab 7 \tab 2 \tab 2 \tab 3 \tab 5 \tab 8 \tab 8 \tab 3 \tab 2 \tab 2 \tab 3 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{9997} uncodeable
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
