% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_07.R
\name{paoccindv}
\alias{paoccindv}
\title{paoccindv}
\source{
General Social Survey https://gss.norc.org
}
\description{
Father's occ/ind verbatim retrieved
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab no verbatims for pa \tab verbatims for pa \tab Total \cr
   1972 \tab 911 \tab 70 \tab 632 \tab 1613 \cr
   1973 \tab 270 \tab 73 \tab 1161 \tab 1504 \cr
   1974 \tab 267 \tab 136 \tab 1081 \tab 1484 \cr
   1975 \tab 514 \tab 87 \tab 889 \tab 1490 \cr
   1976 \tab 269 \tab 148 \tab 1082 \tab 1499 \cr
   1977 \tab 713 \tab 100 \tab 717 \tab 1530 \cr
   1978 \tab 14 \tab 182 \tab 1336 \tab 1532 \cr
   1980 \tab 175 \tab 174 \tab 1119 \tab 1468 \cr
   1982 \tab 164 \tab 268 \tab 1428 \tab 1860 \cr
   1983 \tab 124 \tab 172 \tab 1303 \tab 1599 \cr
   1984 \tab 15 \tab 168 \tab 1290 \tab 1473 \cr
   1985 \tab 238 \tab 147 \tab 1149 \tab 1534 \cr
   1986 \tab 45 \tab 186 \tab 1239 \tab 1470 \cr
   1987 \tab 308 \tab 222 \tab 1289 \tab 1819 \cr
   1988 \tab 21 \tab 178 \tab 1282 \tab 1481 \cr
   1989 \tab 117 \tab 175 \tab 1245 \tab 1537 \cr
   1990 \tab 14 \tab 190 \tab 1168 \tab 1372 \cr
   1991 \tab 18 \tab 208 \tab 1291 \tab 1517 \cr
   1993 \tab 61 \tab 207 \tab 1338 \tab 1606 \cr
   1994 \tab 54 \tab 485 \tab 2453 \tab 2992 \cr
   1996 \tab 21 \tab 487 \tab 2396 \tab 2904 \cr
   1998 \tab 31 \tab 470 \tab 2331 \tab 2832 \cr
   2000 \tab - \tab 550 \tab 2267 \tab 2817 \cr
   2002 \tab - \tab 549 \tab 2216 \tab 2765 \cr
   2004 \tab - \tab 570 \tab 2242 \tab 2812 \cr
   2006 \tab - \tab 2067 \tab 2443 \tab 4510 \cr
   2008 \tab - \tab 344 \tab 1679 \tab 2023 \cr
   2010 \tab - \tab 355 \tab 1689 \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab 3544 \cr
   Total \tab 21667 \tab 8968 \tab 41755 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} no verbatims for pa
\item \code{1} verbatims for pa
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
