% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_07.R
\name{paoccstatus}
\alias{paoccstatus}
\title{Paocc10 coding status}
\source{
General Social Survey https://gss.norc.org
}
\description{
paoccstatus
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab found, coded \tab found, iap \tab found, no answer \tab not found, coded \tab not found, iap \tab not found, no answer \tab not found, no code assigned \tab iap \tab Total \cr
   1972 \tab 584 \tab 98 \tab 20 \tab 623 \tab 100 \tab 37 \tab 151 \tab - \tab 1613 \cr
   1973 \tab 1088 \tab 126 \tab 20 \tab 162 \tab 25 \tab 8 \tab 75 \tab - \tab 1504 \cr
   1974 \tab 1004 \tab 148 \tab 65 \tab 186 \tab 34 \tab 1 \tab 46 \tab - \tab 1484 \cr
   1975 \tab 869 \tab 96 \tab 11 \tab 332 \tab 57 \tab 8 \tab 117 \tab - \tab 1490 \cr
   1976 \tab 1059 \tab 154 \tab 17 \tab 178 \tab 33 \tab 7 \tab 51 \tab - \tab 1499 \cr
   1977 \tab 700 \tab 97 \tab 20 \tab 465 \tab 96 \tab 10 \tab 142 \tab - \tab 1530 \cr
   1978 \tab 1311 \tab 192 \tab 15 \tab 1 \tab 13 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1088 \tab 198 \tab 7 \tab 98 \tab 32 \tab - \tab 45 \tab - \tab 1468 \cr
   1982 \tab 1386 \tab 262 \tab 48 \tab 92 \tab 34 \tab 9 \tab 29 \tab - \tab 1860 \cr
   1983 \tab 1280 \tab 182 \tab 13 \tab 79 \tab 25 \tab - \tab 20 \tab - \tab 1599 \cr
   1984 \tab 1249 \tab 194 \tab 15 \tab 3 \tab 8 \tab 1 \tab 3 \tab - \tab 1473 \cr
   1985 \tab 1128 \tab 160 \tab 8 \tab 158 \tab 34 \tab 2 \tab 44 \tab - \tab 1534 \cr
   1986 \tab 1202 \tab 200 \tab 23 \tab 15 \tab 22 \tab 1 \tab 7 \tab - \tab 1470 \cr
   1987 \tab 1255 \tab 237 \tab 19 \tab 182 \tab 54 \tab 9 \tab 63 \tab - \tab 1819 \cr
   1988 \tab 1244 \tab 197 \tab 19 \tab 10 \tab 10 \tab - \tab 1 \tab - \tab 1481 \cr
   1989 \tab 1208 \tab 196 \tab 16 \tab 82 \tab 29 \tab 2 \tab 4 \tab - \tab 1537 \cr
   1990 \tab 1135 \tab 207 \tab 16 \tab 4 \tab 10 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1248 \tab 211 \tab 40 \tab 3 \tab 9 \tab 4 \tab 2 \tab - \tab 1517 \cr
   1993 \tab 1325 \tab 182 \tab 38 \tab 37 \tab 15 \tab 4 \tab 5 \tab - \tab 1606 \cr
   1994 \tab 2402 \tab 497 \tab 39 \tab 39 \tab 11 \tab 2 \tab 2 \tab - \tab 2992 \cr
   1996 \tab 2324 \tab 495 \tab 64 \tab 12 \tab 7 \tab - \tab 2 \tab - \tab 2904 \cr
   1998 \tab 2250 \tab 488 \tab 63 \tab 13 \tab 15 \tab 2 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 2241 \tab 518 \tab 58 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2176 \tab 526 \tab 63 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2209 \tab 514 \tab 89 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2403 \tab 2049 \tab 58 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1625 \tab 354 \tab 44 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1642 \tab 354 \tab 48 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 40635 \tab 9132 \tab 956 \tab 2774 \tab 673 \tab 107 \tab 810 \tab 17303 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} found, iap
\item \code{2} found, no answer
\item \code{3} found, coded
\item \code{4} not found, iap
\item \code{5} not found, no answer
\item \code{6} not found, coded
\item \code{7} not found, no code assigned
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
