% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_07.R
\name{papres10}
\alias{papres10}
\title{papres10}
\source{
General Social Survey https://gss.norc.org
}
\description{
Father's occupational prestige score (2010)
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 17 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 69 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 80 \tab iap \tab 54 \tab 55 \tab 56 \tab 57 \tab 68 \tab 16 \tab 18 \tab 20 \tab 67 \tab 76 \tab Total \cr
   1972 \tab 1 \tab 2 \tab 4 \tab 1 \tab 22 \tab 29 \tab 10 \tab 60 \tab 27 \tab 22 \tab 4 \tab 22 \tab 22 \tab 17 \tab 13 \tab 79 \tab 22 \tab 9 \tab 39 \tab 45 \tab 55 \tab 3 \tab 28 \tab 18 \tab 40 \tab 64 \tab 308 \tab 31 \tab 7 \tab 25 \tab 35 \tab 10 \tab 2 \tab 12 \tab 3 \tab 15 \tab 23 \tab 2 \tab 8 \tab 3 \tab 7 \tab 3 \tab 16 \tab 3 \tab 7 \tab 2 \tab 7 \tab 8 \tab 1 \tab 3 \tab 8 \tab 406 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1 \tab 2 \tab 4 \tab 1 \tab 11 \tab 24 \tab 7 \tab 87 \tab 25 \tab 15 \tab 9 \tab 30 \tab 24 \tab 12 \tab 17 \tab 96 \tab 29 \tab 9 \tab 58 \tab 41 \tab 57 \tab 8 \tab 29 \tab 14 \tab 55 \tab 62 \tab 246 \tab 25 \tab 12 \tab 38 \tab 49 \tab 3 \tab 6 \tab 11 \tab 2 \tab 11 \tab 26 \tab 4 \tab 5 \tab 4 \tab 2 \tab 7 \tab 23 \tab 3 \tab 7 \tab - \tab 10 \tab 11 \tab 3 \tab 2 \tab 4 \tab 254 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 7 \tab 2 \tab 3 \tab 12 \tab 25 \tab 7 \tab 68 \tab 24 \tab 7 \tab 5 \tab 24 \tab 34 \tab 13 \tab 8 \tab 83 \tab 32 \tab 7 \tab 47 \tab 40 \tab 50 \tab 13 \tab 21 \tab 21 \tab 54 \tab 79 \tab 239 \tab 36 \tab 11 \tab 25 \tab 31 \tab 8 \tab 3 \tab 17 \tab 7 \tab 6 \tab 22 \tab 5 \tab 9 \tab 8 \tab 3 \tab 6 \tab 11 \tab 2 \tab 5 \tab 4 \tab 17 \tab 6 \tab 4 \tab 4 \tab 3 \tab 294 \tab 2 \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1484 \cr
   1975 \tab 1 \tab 2 \tab 5 \tab 3 \tab 13 \tab 21 \tab 7 \tab 29 \tab 30 \tab 15 \tab 10 \tab 28 \tab 27 \tab 9 \tab 11 \tab 103 \tab 41 \tab 3 \tab 46 \tab 43 \tab 52 \tab 7 \tab 32 \tab 20 \tab 53 \tab 55 \tab 268 \tab 28 \tab 10 \tab 39 \tab 34 \tab 16 \tab 2 \tab 18 \tab 4 \tab 13 \tab 17 \tab 5 \tab 9 \tab 5 \tab 3 \tab 4 \tab 15 \tab 2 \tab 8 \tab 1 \tab 6 \tab 9 \tab 3 \tab 2 \tab 7 \tab 289 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1 \tab 7 \tab 4 \tab - \tab 18 \tab 33 \tab 8 \tab 51 \tab 33 \tab 14 \tab 7 \tab 23 \tab 29 \tab 14 \tab 15 \tab 104 \tab 23 \tab 7 \tab 62 \tab 32 \tab 55 \tab 1 \tab 30 \tab 7 \tab 47 \tab 64 \tab 262 \tab 39 \tab 13 \tab 42 \tab 44 \tab 11 \tab 2 \tab 17 \tab - \tab 7 \tab 17 \tab 5 \tab 5 \tab 8 \tab 6 \tab 5 \tab 13 \tab 3 \tab 6 \tab 2 \tab 12 \tab 7 \tab 3 \tab 4 \tab 10 \tab 262 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1499 \cr
   1977 \tab - \tab 4 \tab 4 \tab 2 \tab 16 \tab 31 \tab 8 \tab 47 \tab 24 \tab 17 \tab 4 \tab 23 \tab 36 \tab 12 \tab 12 \tab 82 \tab 25 \tab 15 \tab 59 \tab 25 \tab 69 \tab 3 \tab 32 \tab 15 \tab 37 \tab 56 \tab 261 \tab 26 \tab 13 \tab 22 \tab 39 \tab 9 \tab 2 \tab 12 \tab 6 \tab 12 \tab 22 \tab 4 \tab 13 \tab 6 \tab 1 \tab 6 \tab 10 \tab 8 \tab 2 \tab 3 \tab 7 \tab 5 \tab 3 \tab 1 \tab 4 \tab 365 \tab - \tab - \tab 4 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1 \tab - \tab 4 \tab - \tab 21 \tab 26 \tab 6 \tab 49 \tab 20 \tab 19 \tab 11 \tab 27 \tab 31 \tab 11 \tab 14 \tab 98 \tab 37 \tab 8 \tab 85 \tab 44 \tab 71 \tab 3 \tab 31 \tab 15 \tab 67 \tab 70 \tab 229 \tab 26 \tab 19 \tab 47 \tab 44 \tab 8 \tab 2 \tab 15 \tab 4 \tab 16 \tab 24 \tab 6 \tab 10 \tab 10 \tab 5 \tab 5 \tab 11 \tab 9 \tab 8 \tab 1 \tab 8 \tab 17 \tab 2 \tab - \tab 7 \tab 220 \tab 4 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab 4 \tab - \tab 17 \tab 25 \tab 10 \tab 37 \tab 27 \tab 7 \tab 4 \tab 29 \tab 20 \tab 8 \tab 14 \tab 83 \tab 37 \tab 9 \tab 83 \tab 42 \tab 55 \tab 3 \tab 30 \tab 21 \tab 41 \tab 59 \tab 219 \tab 28 \tab 18 \tab 40 \tab 37 \tab 8 \tab 5 \tab 20 \tab 1 \tab 14 \tab 20 \tab 5 \tab 14 \tab 9 \tab 4 \tab 2 \tab 9 \tab 10 \tab 5 \tab - \tab 7 \tab 12 \tab 9 \tab 5 \tab 9 \tab 282 \tab 2 \tab 1 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 3 \tab 6 \tab 2 \tab 31 \tab 41 \tab 7 \tab 65 \tab 39 \tab 21 \tab 8 \tab 39 \tab 41 \tab 10 \tab 13 \tab 138 \tab 30 \tab 8 \tab 66 \tab 59 \tab 58 \tab 4 \tab 45 \tab 21 \tab 55 \tab 97 \tab 261 \tab 32 \tab 13 \tab 36 \tab 47 \tab 6 \tab 2 \tab 24 \tab 4 \tab 24 \tab 10 \tab 4 \tab 11 \tab 11 \tab 8 \tab 10 \tab 20 \tab 5 \tab 4 \tab - \tab 12 \tab 14 \tab 3 \tab 2 \tab 5 \tab 382 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1860 \cr
   1983 \tab 1 \tab 1 \tab 5 \tab 2 \tab 26 \tab 27 \tab 10 \tab 47 \tab 27 \tab 16 \tab 6 \tab 30 \tab 27 \tab 17 \tab 21 \tab 108 \tab 42 \tab 7 \tab 78 \tab 50 \tab 56 \tab 9 \tab 42 \tab 18 \tab 53 \tab 77 \tab 190 \tab 38 \tab 17 \tab 41 \tab 59 \tab 3 \tab 3 \tab 20 \tab 8 \tab 8 \tab 23 \tab 4 \tab 10 \tab 9 \tab 6 \tab 8 \tab 17 \tab 10 \tab 20 \tab - \tab 16 \tab 13 \tab 11 \tab 4 \tab 8 \tab 240 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1 \tab 1 \tab 2 \tab 2 \tab 14 \tab 28 \tab 3 \tab 39 \tab 30 \tab 9 \tab 3 \tab 26 \tab 22 \tab 18 \tab 17 \tab 110 \tab 32 \tab 6 \tab 82 \tab 49 \tab 41 \tab 4 \tab 30 \tab 29 \tab 41 \tab 50 \tab 236 \tab 29 \tab 17 \tab 43 \tab 52 \tab 8 \tab - \tab 22 \tab 6 \tab 10 \tab 17 \tab 5 \tab 12 \tab 13 \tab 6 \tab 7 \tab 13 \tab 5 \tab 7 \tab 1 \tab 17 \tab 13 \tab 5 \tab 3 \tab 3 \tab 221 \tab 3 \tab - \tab 5 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 2 \tab 4 \tab 1 \tab 13 \tab 23 \tab 11 \tab 40 \tab 39 \tab 7 \tab 3 \tab 29 \tab 24 \tab 13 \tab 10 \tab 96 \tab 29 \tab 11 \tab 83 \tab 45 \tab 41 \tab 8 \tab 26 \tab 20 \tab 46 \tab 70 \tab 229 \tab 25 \tab 14 \tab 46 \tab 61 \tab 6 \tab 3 \tab 22 \tab 8 \tab 20 \tab 21 \tab 4 \tab 10 \tab 14 \tab 7 \tab 8 \tab 17 \tab 8 \tab 11 \tab 5 \tab 17 \tab 9 \tab 5 \tab 4 \tab 10 \tab 248 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1534 \cr
   1986 \tab 1 \tab 1 \tab 6 \tab 1 \tab 15 \tab 25 \tab 9 \tab 53 \tab 24 \tab 6 \tab 8 \tab 37 \tab 25 \tab 8 \tab 19 \tab 95 \tab 21 \tab 7 \tab 82 \tab 43 \tab 42 \tab 8 \tab 30 \tab 22 \tab 42 \tab 72 \tab 215 \tab 23 \tab 7 \tab 41 \tab 52 \tab 10 \tab 1 \tab 25 \tab 4 \tab 12 \tab 17 \tab 1 \tab 8 \tab 10 \tab 5 \tab 9 \tab 16 \tab 8 \tab 4 \tab 1 \tab 13 \tab 9 \tab 7 \tab 6 \tab 5 \tab 253 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 2 \tab 7 \tab - \tab 18 \tab 29 \tab 13 \tab 70 \tab 38 \tab 17 \tab 6 \tab 24 \tab 26 \tab 14 \tab 24 \tab 139 \tab 20 \tab 9 \tab 75 \tab 47 \tab 50 \tab 7 \tab 35 \tab 22 \tab 61 \tab 61 \tab 262 \tab 30 \tab 15 \tab 39 \tab 48 \tab 9 \tab 2 \tab 25 \tab 2 \tab 19 \tab 24 \tab 4 \tab 13 \tab 15 \tab 12 \tab 10 \tab 13 \tab 9 \tab 9 \tab 2 \tab 12 \tab 17 \tab 8 \tab 2 \tab 12 \tab 382 \tab 1 \tab 1 \tab 5 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1819 \cr
   1988 \tab 1 \tab 2 \tab 4 \tab 2 \tab 17 \tab 26 \tab 9 \tab 42 \tab 24 \tab 12 \tab 3 \tab 33 \tab 23 \tab 12 \tab 11 \tab 113 \tab 26 \tab 7 \tab 90 \tab 41 \tab 34 \tab 11 \tab 43 \tab 24 \tab 40 \tab 68 \tab 195 \tab 28 \tab 17 \tab 39 \tab 52 \tab 7 \tab 4 \tab 20 \tab - \tab 14 \tab 23 \tab 7 \tab 8 \tab 8 \tab 8 \tab 10 \tab 14 \tab 6 \tab 11 \tab 4 \tab 19 \tab 16 \tab 7 \tab 4 \tab 8 \tab 227 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1481 \cr
   1989 \tab 2 \tab 2 \tab 1 \tab 2 \tab 11 \tab 23 \tab 11 \tab 51 \tab 32 \tab 6 \tab 7 \tab 24 \tab 18 \tab 8 \tab 12 \tab 123 \tab 25 \tab 6 \tab 68 \tab 43 \tab 50 \tab 7 \tab 35 \tab 22 \tab 35 \tab 75 \tab 207 \tab 26 \tab 19 \tab 57 \tab 60 \tab 8 \tab 1 \tab 20 \tab 5 \tab 21 \tab 21 \tab 8 \tab 13 \tab 8 \tab 9 \tab 13 \tab 18 \tab 6 \tab 11 \tab 1 \tab 15 \tab 17 \tab 11 \tab 4 \tab 6 \tab 247 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 1 \tab 3 \tab 1 \tab 11 \tab 26 \tab 8 \tab 31 \tab 34 \tab 5 \tab 3 \tab 27 \tab 15 \tab 12 \tab 20 \tab 111 \tab 20 \tab 13 \tab 100 \tab 37 \tab 24 \tab 4 \tab 28 \tab 23 \tab 50 \tab 64 \tab 162 \tab 26 \tab 34 \tab 22 \tab 49 \tab 7 \tab 4 \tab 20 \tab 3 \tab 12 \tab 12 \tab 6 \tab 8 \tab 4 \tab 6 \tab 11 \tab 11 \tab 2 \tab 11 \tab 2 \tab 15 \tab 10 \tab 6 \tab 6 \tab 6 \tab 233 \tab 2 \tab - \tab 7 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1372 \cr
   1991 \tab - \tab 2 \tab 2 \tab 2 \tab 22 \tab 23 \tab 5 \tab 48 \tab 25 \tab 6 \tab 10 \tab 22 \tab 25 \tab 14 \tab 15 \tab 109 \tab 32 \tab 13 \tab 89 \tab 45 \tab 35 \tab 8 \tab 35 \tab 22 \tab 36 \tab 82 \tab 196 \tab 30 \tab 16 \tab 43 \tab 39 \tab 5 \tab 4 \tab 22 \tab 2 \tab 11 \tab 19 \tab 7 \tab 7 \tab 17 \tab 8 \tab 8 \tab 13 \tab 3 \tab 10 \tab 2 \tab 13 \tab 23 \tab 3 \tab 3 \tab 10 \tab 266 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1517 \cr
   1993 \tab 1 \tab 1 \tab 5 \tab - \tab 18 \tab 26 \tab 8 \tab 46 \tab 23 \tab 13 \tab 4 \tab 52 \tab 26 \tab 16 \tab 8 \tab 113 \tab 30 \tab 13 \tab 93 \tab 61 \tab 33 \tab 2 \tab 33 \tab 20 \tab 42 \tab 70 \tab 205 \tab 27 \tab 15 \tab 40 \tab 48 \tab 10 \tab 3 \tab 42 \tab 2 \tab 21 \tab 28 \tab 6 \tab 9 \tab 10 \tab 11 \tab 18 \tab 12 \tab 10 \tab 12 \tab 1 \tab 22 \tab 16 \tab 8 \tab 3 \tab 13 \tab 244 \tab 2 \tab 3 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1606 \cr
   1994 \tab 2 \tab 8 \tab 6 \tab - \tab 48 \tab 54 \tab 14 \tab 74 \tab 61 \tab 16 \tab 15 \tab 49 \tab 48 \tab 29 \tab 28 \tab 228 \tab 60 \tab 25 \tab 148 \tab 106 \tab 66 \tab 7 \tab 65 \tab 34 \tab 76 \tab 134 \tab 312 \tab 52 \tab 29 \tab 70 \tab 82 \tab 15 \tab 7 \tab 47 \tab 8 \tab 37 \tab 44 \tab 15 \tab 13 \tab 26 \tab 20 \tab 28 \tab 40 \tab 24 \tab 19 \tab 14 \tab 44 \tab 29 \tab 15 \tab 10 \tab 11 \tab 551 \tab 2 \tab 3 \tab 12 \tab 7 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2992 \cr
   1996 \tab 2 \tab 2 \tab 7 \tab 1 \tab 33 \tab 44 \tab 16 \tab 70 \tab 68 \tab 18 \tab 17 \tab 58 \tab 45 \tab 17 \tab 18 \tab 207 \tab 52 \tab 19 \tab 156 \tab 93 \tab 64 \tab 14 \tab 58 \tab 36 \tab 59 \tab 124 \tab 259 \tab 52 \tab 44 \tab 100 \tab 104 \tab 19 \tab 9 \tab 50 \tab 5 \tab 28 \tab 49 \tab 18 \tab 11 \tab 24 \tab 25 \tab 31 \tab 36 \tab 16 \tab 18 \tab 9 \tab 39 \tab 36 \tab 10 \tab 10 \tab 15 \tab 568 \tab 3 \tab 3 \tab 8 \tab 5 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2904 \cr
   1998 \tab 1 \tab 2 \tab 5 \tab 2 \tab 36 \tab 33 \tab 5 \tab 64 \tab 47 \tab 19 \tab 11 \tab 56 \tab 33 \tab 20 \tab 33 \tab 211 \tab 43 \tab 20 \tab 139 \tab 119 \tab 72 \tab 14 \tab 45 \tab 41 \tab 72 \tab 115 \tab 254 \tab 49 \tab 35 \tab 85 \tab 96 \tab 24 \tab 7 \tab 57 \tab 7 \tab 22 \tab 48 \tab 12 \tab 14 \tab 29 \tab 28 \tab 27 \tab 30 \tab 21 \tab 27 \tab 2 \tab 28 \tab 34 \tab 22 \tab 4 \tab 17 \tab 569 \tab 2 \tab 1 \tab 9 \tab 2 \tab 2 \tab 4 \tab - \tab 2 \tab 4 \tab - \tab 2832 \cr
   2000 \tab 1 \tab 3 \tab 7 \tab 4 \tab 38 \tab 49 \tab 13 \tab 69 \tab 56 \tab 24 \tab 13 \tab 63 \tab 45 \tab 16 \tab 29 \tab 237 \tab 48 \tab 19 \tab 122 \tab 84 \tab 52 \tab 11 \tab 62 \tab 39 \tab 62 \tab 129 \tab 267 \tab 49 \tab 31 \tab 72 \tab 80 \tab 12 \tab 9 \tab 58 \tab 12 \tab 34 \tab 40 \tab 11 \tab 14 \tab 23 \tab 19 \tab 21 \tab 34 \tab 13 \tab 16 \tab 5 \tab 29 \tab 39 \tab 16 \tab 9 \tab 13 \tab 576 \tab 3 \tab 2 \tab 9 \tab 2 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 2817 \cr
   2002 \tab 2 \tab 3 \tab 2 \tab - \tab 29 \tab 43 \tab 14 \tab 58 \tab 55 \tab 24 \tab 8 \tab 58 \tab 39 \tab 35 \tab 35 \tab 216 \tab 33 \tab 15 \tab 132 \tab 96 \tab 41 \tab 9 \tab 50 \tab 40 \tab 78 \tab 121 \tab 222 \tab 39 \tab 42 \tab 87 \tab 89 \tab 10 \tab 8 \tab 54 \tab 13 \tab 28 \tab 60 \tab 13 \tab 26 \tab 25 \tab 21 \tab 23 \tab 28 \tab 12 \tab 20 \tab 8 \tab 27 \tab 24 \tab 16 \tab 8 \tab 13 \tab 589 \tab 6 \tab 3 \tab 6 \tab 6 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2765 \cr
   2004 \tab - \tab 5 \tab 3 \tab - \tab 35 \tab 27 \tab 12 \tab 71 \tab 74 \tab 23 \tab 6 \tab 67 \tab 39 \tab 23 \tab 29 \tab 226 \tab 47 \tab 22 \tab 130 \tab 89 \tab 70 \tab 13 \tab 41 \tab 38 \tab 78 \tab 116 \tab 189 \tab 50 \tab 36 \tab 61 \tab 96 \tab 10 \tab 12 \tab 57 \tab 5 \tab 27 \tab 54 \tab 17 \tab 26 \tab 22 \tab 28 \tab 27 \tab 21 \tab 20 \tab 25 \tab 6 \tab 30 \tab 31 \tab 23 \tab 11 \tab 20 \tab 603 \tab 2 \tab 2 \tab 10 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2812 \cr
   2006 \tab - \tab 2 \tab 7 \tab 3 \tab 38 \tab 50 \tab 9 \tab 87 \tab 75 \tab 32 \tab 9 \tab 58 \tab 43 \tab 25 \tab 33 \tab 240 \tab 52 \tab 18 \tab 138 \tab 99 \tab 52 \tab 7 \tab 59 \tab 48 \tab 72 \tab 132 \tab 223 \tab 58 \tab 38 \tab 98 \tab 80 \tab 19 \tab 6 \tab 47 \tab 7 \tab 34 \tab 48 \tab 29 \tab 38 \tab 24 \tab 16 \tab 20 \tab 37 \tab 20 \tab 24 \tab 11 \tab 34 \tab 35 \tab 12 \tab 9 \tab 20 \tab 2107 \tab 8 \tab 5 \tab 6 \tab 3 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab 4510 \cr
   2008 \tab 2 \tab 4 \tab 7 \tab 1 \tab 29 \tab 28 \tab 9 \tab 39 \tab 47 \tab 19 \tab 7 \tab 33 \tab 23 \tab 28 \tab 16 \tab 151 \tab 43 \tab 14 \tab 82 \tab 71 \tab 52 \tab 9 \tab 44 \tab 39 \tab 34 \tab 89 \tab 144 \tab 51 \tab 22 \tab 74 \tab 60 \tab 18 \tab 4 \tab 45 \tab 6 \tab 20 \tab 38 \tab 13 \tab 31 \tab 18 \tab 13 \tab 12 \tab 18 \tab 8 \tab 20 \tab 2 \tab 21 \tab 28 \tab 9 \tab 6 \tab 10 \tab 398 \tab 1 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 1 \tab 2 \tab 1 \tab 2 \tab 30 \tab 30 \tab 15 \tab 44 \tab 51 \tab 11 \tab 9 \tab 48 \tab 22 \tab 21 \tab 12 \tab 152 \tab 39 \tab 16 \tab 96 \tab 90 \tab 34 \tab 7 \tab 52 \tab 38 \tab 47 \tab 93 \tab 128 \tab 38 \tab 30 \tab 59 \tab 62 \tab 12 \tab 2 \tab 40 \tab 11 \tab 12 \tab 37 \tab 12 \tab 19 \tab 22 \tab 16 \tab 20 \tab 12 \tab 10 \tab 18 \tab 3 \tab 28 \tab 24 \tab 19 \tab 8 \tab 20 \tab 402 \tab 4 \tab 2 \tab 4 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2044 \cr
   2012 \tab - \tab 5 \tab 2 \tab 3 \tab 36 \tab 37 \tab 5 \tab 54 \tab 65 \tab 12 \tab 7 \tab 38 \tab 19 \tab 17 \tab 17 \tab 165 \tab 22 \tab 13 \tab 83 \tab 67 \tab 47 \tab 9 \tab 32 \tab 20 \tab 47 \tab 83 \tab 128 \tab 43 \tab 12 \tab 46 \tab 41 \tab 13 \tab 7 \tab 40 \tab 4 \tab 29 \tab 28 \tab 11 \tab 8 \tab 62 \tab 20 \tab 19 \tab 21 \tab 8 \tab 21 \tab - \tab 20 \tab 40 \tab 21 \tab 5 \tab 22 \tab 385 \tab 1 \tab 5 \tab 3 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 1 \tab 4 \tab - \tab 35 \tab 37 \tab 14 \tab 64 \tab 71 \tab 9 \tab 5 \tab 57 \tab 28 \tab 22 \tab 23 \tab 202 \tab 33 \tab 21 \tab 116 \tab 81 \tab 53 \tab 16 \tab 54 \tab 48 \tab 64 \tab 125 \tab 149 \tab 48 \tab 33 \tab 57 \tab 84 \tab 15 \tab 11 \tab 48 \tab 11 \tab 17 \tab 52 \tab 9 \tab 46 \tab 22 \tab 19 \tab 22 \tab 22 \tab 17 \tab 25 \tab 6 \tab 29 \tab 20 \tab 19 \tab 6 \tab 20 \tab 524 \tab 4 \tab 4 \tab 5 \tab 6 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 2538 \cr
   2016 \tab 2 \tab 2 \tab 8 \tab 2 \tab 58 \tab 38 \tab 13 \tab 83 \tab 83 \tab 21 \tab 12 \tab 74 \tab 34 \tab 33 \tab 33 \tab 225 \tab 42 \tab 21 \tab 115 \tab 93 \tab 63 \tab 13 \tab 60 \tab 44 \tab 58 \tab 114 \tab 142 \tab 38 \tab 44 \tab 60 \tab 92 \tab 26 \tab 5 \tab 62 \tab 12 \tab 30 \tab 65 \tab 12 \tab 26 \tab 27 \tab 27 \tab 28 \tab 23 \tab 16 \tab 37 \tab 6 \tab 23 \tab 24 \tab 16 \tab 6 \tab 17 \tab 625 \tab 3 \tab 5 \tab 11 \tab 8 \tab 2 \tab - \tab - \tab 1 \tab 4 \tab - \tab 2867 \cr
   2018 \tab 1 \tab 5 \tab 3 \tab - \tab 40 \tab 37 \tab 17 \tab 57 \tab 56 \tab 16 \tab 9 \tab 68 \tab 33 \tab 24 \tab 14 \tab 168 \tab 27 \tab 17 \tab 97 \tab 52 \tab 39 \tab 8 \tab 48 \tab 30 \tab 60 \tab 102 \tab 135 \tab 47 \tab 59 \tab 83 \tab 71 \tab 16 \tab 5 \tab 44 \tab 10 \tab 30 \tab 42 \tab 12 \tab 33 \tab 21 \tab 19 \tab 23 \tab 23 \tab 8 \tab 12 \tab 3 \tab 34 \tab 21 \tab 13 \tab 6 \tab 19 \tab 506 \tab 4 \tab 1 \tab 5 \tab 5 \tab 5 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 2348 \cr
   2021 \tab 4 \tab 4 \tab 9 \tab 3 \tab 56 \tab 55 \tab 16 \tab 61 \tab 100 \tab 20 \tab 10 \tab 85 \tab 61 \tab 44 \tab 34 \tab 350 \tab 64 \tab 36 \tab 203 \tab 196 \tab 85 \tab 19 \tab 80 \tab 71 \tab 120 \tab 176 \tab 170 \tab 59 \tab 65 \tab 116 \tab 110 \tab 20 \tab 14 \tab 84 \tab 16 \tab 35 \tab 116 \tab 29 \tab 92 \tab 25 \tab 38 \tab 38 \tab 39 \tab 34 \tab 57 \tab 14 \tab 63 \tab 26 \tab 31 \tab 18 \tab 31 \tab 683 \tab 14 \tab 2 \tab 10 \tab 8 \tab 3 \tab 1 \tab 1 \tab - \tab 8 \tab - \tab 4032 \cr
   2022 \tab - \tab 7 \tab 8 \tab 1 \tab 45 \tab 66 \tab 10 \tab 67 \tab 113 \tab 15 \tab 13 \tab 88 \tab 39 \tab 26 \tab 31 \tab 295 \tab 47 \tab 40 \tab 130 \tab 143 \tab 66 \tab 20 \tab 67 \tab 37 \tab 87 \tab 160 \tab 123 \tab 54 \tab 57 \tab 94 \tab 77 \tab 25 \tab 11 \tab 83 \tab 11 \tab 33 \tab 104 \tab 30 \tab 86 \tab 29 \tab 32 \tab 36 \tab 30 \tab 25 \tab 35 \tab 7 \tab 41 \tab 23 \tab 22 \tab 14 \tab 33 \tab 770 \tab 10 \tab 7 \tab 14 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 3544 \cr
   Total \tab 30 \tab 97 \tab 155 \tab 47 \tab 912 \tab 1139 \tab 339 \tab 1922 \tab 1567 \tab 513 \tab 266 \tab 1449 \tab 1046 \tab 618 \tab 659 \tab 5256 \tab 1205 \tab 483 \tab 3272 \tab 2311 \tab 1784 \tab 289 \tab 1432 \tab 977 \tab 1909 \tab 3108 \tab 7235 \tab 1280 \tab 864 \tab 1887 \tab 2104 \tab 406 \tab 168 \tab 1200 \tab 209 \tab 682 \tab 1211 \tab 335 \tab 665 \tab 571 \tab 458 \tab 525 \tab 686 \tab 364 \tab 532 \tab 128 \tab 735 \tab 666 \tab 366 \tab 192 \tab 419 \tab 15171 \tab 97 \tab 62 \tab 172 \tab 94 \tab 38 \tab 14 \tab 17 \tab 13 \tab 38 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
