% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{papres105plus}
\alias{papres105plus}
\title{Father's occupational prestige score using threshold method (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
papres105plus
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 58 \tab 59 \tab 6 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 69 \tab 72 \tab 73 \tab 75 \tab 76 \tab 78 \tab 8 \tab 80 \tab 81 \tab 82 \tab 83 \tab 84 \tab 85 \tab 86 \tab 87 \tab 88 \tab 89 \tab 90 \tab 91 \tab 92 \tab 94 \tab 95 \tab iap \tab 43 \tab 57 \tab 68 \tab 70 \tab 71 \tab 77 \tab 9 \tab 93 \tab 30 \tab 7 \tab 74 \tab 97 \tab 5 \tab 79 \tab 67 \tab not available in this release \tab Total \cr
   1972 \tab 4 \tab 7 \tab 10 \tab 4 \tab 3 \tab 18 \tab 23 \tab 7 \tab 71 \tab 31 \tab 3 \tab 14 \tab 3 \tab 23 \tab 9 \tab 3 \tab 32 \tab 19 \tab 12 \tab 59 \tab 5 \tab 3 \tab 12 \tab 5 \tab 7 \tab 11 \tab 23 \tab 18 \tab 17 \tab 10 \tab 1 \tab 43 \tab 5 \tab 7 \tab 17 \tab 3 \tab 7 \tab 50 \tab 41 \tab 4 \tab 1 \tab 17 \tab 12 \tab 12 \tab 4 \tab 305 \tab 1 \tab 1 \tab 4 \tab 14 \tab 30 \tab 12 \tab 3 \tab 12 \tab 2 \tab 6 \tab 2 \tab 2 \tab 4 \tab 7 \tab 10 \tab 2 \tab 10 \tab 1 \tab 2 \tab 1 \tab 6 \tab 13 \tab 11 \tab 2 \tab 4 \tab 3 \tab 7 \tab 11 \tab 3 \tab 18 \tab 8 \tab 406 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1 \tab 8 \tab 2 \tab 3 \tab 1 \tab 12 \tab 27 \tab 2 \tab 101 \tab 17 \tab 5 \tab 22 \tab 5 \tab 21 \tab 16 \tab 11 \tab 27 \tab 13 \tab 28 \tab 65 \tab 4 \tab 4 \tab 10 \tab 5 \tab 11 \tab 18 \tab 19 \tab 11 \tab 38 \tab 7 \tab 2 \tab 47 \tab 7 \tab 9 \tab 10 \tab 5 \tab 8 \tab 52 \tab 32 \tab 6 \tab 2 \tab 12 \tab 17 \tab 24 \tab 4 \tab 240 \tab 1 \tab 1 \tab 3 \tab 9 \tab 40 \tab 26 \tab 1 \tab 11 \tab 2 \tab 5 \tab - \tab - \tab 9 \tab 4 \tab 9 \tab 2 \tab 14 \tab 3 \tab 2 \tab 1 \tab 5 \tab 16 \tab 12 \tab 3 \tab - \tab 4 \tab 14 \tab 11 \tab 4 \tab 17 \tab 9 \tab 254 \tab 13 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 8 \tab 8 \tab - \tab 7 \tab 12 \tab 18 \tab 4 \tab 84 \tab 14 \tab 10 \tab 9 \tab 5 \tab 18 \tab 17 \tab 7 \tab 16 \tab 18 \tab 26 \tab 51 \tab 5 \tab 13 \tab 12 \tab 9 \tab 12 \tab 11 \tab 15 \tab 8 \tab 28 \tab 14 \tab - \tab 37 \tab 3 \tab 8 \tab 15 \tab 5 \tab 5 \tab 56 \tab 54 \tab 10 \tab 4 \tab 19 \tab 14 \tab 23 \tab 2 \tab 237 \tab 1 \tab 2 \tab 2 \tab 11 \tab 24 \tab 15 \tab 4 \tab 6 \tab 3 \tab 7 \tab 5 \tab 9 \tab 5 \tab 8 \tab 11 \tab 7 \tab 3 \tab 3 \tab 1 \tab - \tab 4 \tab 21 \tab 6 \tab 2 \tab 3 \tab 3 \tab 12 \tab 13 \tab 4 \tab 20 \tab 4 \tab 294 \tab 8 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1 \tab 7 \tab 4 \tab 3 \tab 3 \tab 12 \tab 21 \tab 3 \tab 43 \tab 29 \tab 7 \tab 19 \tab 7 \tab 21 \tab 13 \tab 6 \tab 22 \tab 28 \tab 15 \tab 74 \tab 5 \tab 2 \tab 4 \tab 9 \tab 16 \tab 15 \tab 17 \tab 12 \tab 30 \tab 9 \tab - \tab 46 \tab 6 \tab 7 \tab 15 \tab 4 \tab 4 \tab 50 \tab 40 \tab 5 \tab 2 \tab 19 \tab 17 \tab 7 \tab 1 \tab 270 \tab 5 \tab 2 \tab 8 \tab 11 \tab 31 \tab 18 \tab 1 \tab 11 \tab 4 \tab 11 \tab 1 \tab 3 \tab 7 \tab 8 \tab 7 \tab 2 \tab 7 \tab 8 \tab 1 \tab 1 \tab 7 \tab 14 \tab 9 \tab 3 \tab 1 \tab 2 \tab 10 \tab 7 \tab 1 \tab 18 \tab 11 \tab 289 \tab 14 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1 \tab 4 \tab 8 \tab 2 \tab 2 \tab 19 \tab 37 \tab 4 \tab 53 \tab 19 \tab 12 \tab 12 \tab 5 \tab 28 \tab 10 \tab 7 \tab 26 \tab 15 \tab 35 \tab 69 \tab 7 \tab 5 \tab 8 \tab 11 \tab 6 \tab 9 \tab 26 \tab 8 \tab 41 \tab 6 \tab 1 \tab 41 \tab 5 \tab 9 \tab 10 \tab 8 \tab 3 \tab 41 \tab 38 \tab 2 \tab 2 \tab 23 \tab 18 \tab 16 \tab 4 \tab 264 \tab 6 \tab 1 \tab - \tab 19 \tab 35 \tab 24 \tab 3 \tab 8 \tab - \tab 9 \tab - \tab 7 \tab 5 \tab 6 \tab 8 \tab 7 \tab 6 \tab 2 \tab - \tab - \tab 4 \tab 16 \tab 9 \tab - \tab 4 \tab 3 \tab 11 \tab 15 \tab 3 \tab 16 \tab 11 \tab 262 \tab 8 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 2 \tab 2 \tab 9 \tab 1 \tab 3 \tab 14 \tab 30 \tab 5 \tab 57 \tab 23 \tab 10 \tab 19 \tab 3 \tab 24 \tab 17 \tab 4 \tab 22 \tab 20 \tab 16 \tab 55 \tab 5 \tab 5 \tab 13 \tab 8 \tab 22 \tab 18 \tab 24 \tab 4 \tab 15 \tab 8 \tab 2 \tab 53 \tab 10 \tab 6 \tab 15 \tab 4 \tab 8 \tab 43 \tab 29 \tab 10 \tab 5 \tab 11 \tab 17 \tab 4 \tab 3 \tab 258 \tab 2 \tab - \tab 4 \tab 9 \tab 33 \tab 11 \tab 5 \tab 9 \tab 4 \tab 2 \tab 3 \tab 1 \tab 5 \tab 14 \tab 13 \tab 4 \tab 10 \tab 4 \tab 1 \tab - \tab 7 \tab 16 \tab 8 \tab - \tab - \tab 8 \tab 4 \tab 11 \tab 3 \tab 9 \tab 7 \tab 365 \tab 11 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 3 \tab 4 \tab 13 \tab 3 \tab 2 \tab 15 \tab 26 \tab 3 \tab 62 \tab 26 \tab 14 \tab 14 \tab 1 \tab 16 \tab 16 \tab 9 \tab 28 \tab 15 \tab 33 \tab 64 \tab 14 \tab 4 \tab 6 \tab 9 \tab 20 \tab 16 \tab 36 \tab 9 \tab 37 \tab 8 \tab - \tab 55 \tab 6 \tab 10 \tab 13 \tab 3 \tab 7 \tab 69 \tab 29 \tab 11 \tab 5 \tab 15 \tab 22 \tab 26 \tab 3 \tab 231 \tab 6 \tab 1 \tab 1 \tab 12 \tab 40 \tab 22 \tab 4 \tab 7 \tab 5 \tab 3 \tab 1 \tab 5 \tab 7 \tab 7 \tab 12 \tab - \tab 10 \tab 6 \tab 2 \tab 1 \tab 9 \tab 25 \tab 6 \tab 1 \tab - \tab 10 \tab 10 \tab 11 \tab 3 \tab 15 \tab 18 \tab 220 \tab 11 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 2 \tab 8 \tab 8 \tab 3 \tab 1 \tab 13 \tab 22 \tab 3 \tab 55 \tab 17 \tab 7 \tab 11 \tab 9 \tab 18 \tab 12 \tab 8 \tab 19 \tab 17 \tab 20 \tab 65 \tab 3 \tab 5 \tab 9 \tab 7 \tab 15 \tab 17 \tab 24 \tab 13 \tab 47 \tab 8 \tab - \tab 46 \tab 11 \tab 5 \tab 8 \tab 2 \tab 3 \tab 45 \tab 43 \tab 7 \tab 2 \tab 15 \tab 23 \tab 16 \tab 5 \tab 216 \tab 4 \tab - \tab - \tab 13 \tab 32 \tab 20 \tab 3 \tab 15 \tab - \tab 8 \tab 1 \tab 7 \tab 7 \tab 11 \tab 13 \tab - \tab 11 \tab 3 \tab 1 \tab 3 \tab 8 \tab 13 \tab 5 \tab 2 \tab 1 \tab 12 \tab 7 \tab 8 \tab 4 \tab 20 \tab 15 \tab 282 \tab 12 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1 \tab 10 \tab 12 \tab 1 \tab 1 \tab 30 \tab 40 \tab 5 \tab 86 \tab 32 \tab 10 \tab 21 \tab 9 \tab 31 \tab 16 \tab 2 \tab 27 \tab 28 \tab 32 \tab 89 \tab 9 \tab 4 \tab 10 \tab 5 \tab 15 \tab 26 \tab 19 \tab 15 \tab 39 \tab 10 \tab 2 \tab 51 \tab 9 \tab 10 \tab 21 \tab 4 \tab 8 \tab 55 \tab 54 \tab 15 \tab 3 \tab 13 \tab 19 \tab 31 \tab 1 \tab 256 \tab 6 \tab - \tab 5 \tab 11 \tab 39 \tab 16 \tab 7 \tab 6 \tab 5 \tab 11 \tab 2 \tab 4 \tab 16 \tab 12 \tab 4 \tab 1 \tab 17 \tab 1 \tab 1 \tab - \tab 8 \tab 13 \tab 14 \tab 2 \tab - \tab 6 \tab 14 \tab 23 \tab 1 \tab 17 \tab 13 \tab 382 \tab 8 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1860 \cr
   1983 \tab 1 \tab 7 \tab 11 \tab 4 \tab 4 \tab 24 \tab 22 \tab 8 \tab 65 \tab 25 \tab 10 \tab 11 \tab 3 \tab 29 \tab 8 \tab 8 \tab 33 \tab 25 \tab 21 \tab 79 \tab 7 \tab 6 \tab 12 \tab 9 \tab 15 \tab 21 \tab 29 \tab 8 \tab 43 \tab 10 \tab 2 \tab 50 \tab 13 \tab 8 \tab 14 \tab 4 \tab 9 \tab 55 \tab 36 \tab 15 \tab 6 \tab 23 \tab 25 \tab 16 \tab - \tab 188 \tab 8 \tab 1 \tab - \tab 10 \tab 50 \tab 24 \tab 2 \tab 14 \tab 2 \tab 5 \tab 6 \tab 4 \tab 10 \tab 6 \tab 12 \tab 1 \tab 9 \tab 3 \tab 1 \tab 1 \tab 2 \tab 21 \tab 13 \tab 5 \tab 1 \tab 12 \tab 14 \tab 11 \tab 3 \tab 41 \tab 15 \tab 240 \tab 17 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1 \tab 6 \tab 10 \tab 2 \tab 1 \tab 14 \tab 23 \tab 2 \tab 52 \tab 20 \tab 6 \tab 13 \tab 3 \tab 22 \tab 10 \tab 5 \tab 25 \tab 25 \tab 30 \tab 72 \tab 3 \tab 6 \tab 8 \tab 12 \tab 17 \tab 18 \tab 17 \tab 11 \tab 48 \tab 11 \tab 1 \tab 43 \tab 1 \tab 11 \tab 7 \tab 8 \tab 7 \tab 33 \tab 40 \tab 10 \tab 4 \tab 12 \tab 24 \tab 19 \tab 4 \tab 232 \tab 5 \tab 1 \tab 3 \tab 14 \tab 41 \tab 26 \tab 4 \tab 8 \tab 3 \tab 8 \tab 2 \tab 5 \tab 11 \tab 13 \tab 6 \tab 1 \tab 9 \tab 8 \tab - \tab 2 \tab 6 \tab 21 \tab 9 \tab 6 \tab 2 \tab 6 \tab 15 \tab 13 \tab 4 \tab 20 \tab 6 \tab 221 \tab 17 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1 \tab 4 \tab 7 \tab 3 \tab 4 \tab 14 \tab 21 \tab 5 \tab 61 \tab 22 \tab 6 \tab 12 \tab 5 \tab 33 \tab 3 \tab 3 \tab 17 \tab 12 \tab 31 \tab 77 \tab 11 \tab 4 \tab 10 \tab 7 \tab 17 \tab 14 \tab 18 \tab 8 \tab 48 \tab 11 \tab - \tab 36 \tab 4 \tab 8 \tab 10 \tab 3 \tab 7 \tab 54 \tab 37 \tab 14 \tab - \tab 12 \tab 20 \tab 14 \tab 4 \tab 229 \tab 6 \tab 1 \tab - \tab 14 \tab 47 \tab 25 \tab 1 \tab 16 \tab 3 \tab 10 \tab 5 \tab 5 \tab 10 \tab 6 \tab 10 \tab 1 \tab 12 \tab 4 \tab 3 \tab 1 \tab 11 \tab 20 \tab 15 \tab 4 \tab 1 \tab 10 \tab 15 \tab 19 \tab 4 \tab 23 \tab 15 \tab 248 \tab 16 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 8 \tab 13 \tab 4 \tab 6 \tab 18 \tab 20 \tab 3 \tab 74 \tab 10 \tab 9 \tab 15 \tab 4 \tab 20 \tab 11 \tab 5 \tab 19 \tab 21 \tab 34 \tab 56 \tab 9 \tab 3 \tab 9 \tab 8 \tab 18 \tab 8 \tab 31 \tab 7 \tab 42 \tab 8 \tab 1 \tab 37 \tab 9 \tab 6 \tab 16 \tab 2 \tab 7 \tab 44 \tab 37 \tab 9 \tab 1 \tab 16 \tab 23 \tab 27 \tab 7 \tab 200 \tab 3 \tab - \tab 3 \tab 11 \tab 41 \tab 21 \tab 2 \tab 12 \tab 4 \tab 8 \tab 1 \tab 9 \tab 11 \tab 6 \tab 10 \tab - \tab 12 \tab 2 \tab 3 \tab - \tab 6 \tab 14 \tab 11 \tab 3 \tab 2 \tab 8 \tab 13 \tab 12 \tab 2 \tab 25 \tab 7 \tab 253 \tab 13 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 2 \tab 8 \tab 12 \tab 1 \tab 3 \tab 19 \tab 25 \tab 6 \tab 86 \tab 22 \tab 12 \tab 9 \tab 4 \tab 30 \tab 16 \tab 1 \tab 32 \tab 22 \tab 42 \tab 82 \tab 8 \tab 9 \tab 8 \tab 12 \tab 25 \tab 12 \tab 18 \tab 8 \tab 35 \tab 8 \tab - \tab 49 \tab 7 \tab 6 \tab 16 \tab 7 \tab 6 \tab 57 \tab 38 \tab 9 \tab - \tab 14 \tab 21 \tab 14 \tab - \tab 266 \tab 3 \tab - \tab - \tab 18 \tab 38 \tab 25 \tab 5 \tab 11 \tab 4 \tab 6 \tab 1 \tab 6 \tab 9 \tab 12 \tab 8 \tab 2 \tab 14 \tab 4 \tab 1 \tab - \tab 10 \tab 32 \tab 12 \tab 1 \tab 2 \tab 10 \tab 12 \tab 21 \tab 4 \tab 28 \tab 16 \tab 382 \tab 19 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 2 \tab 11 \tab 7 \tab 2 \tab 2 \tab 17 \tab 22 \tab 3 \tab 65 \tab 19 \tab 7 \tab 11 \tab 2 \tab 20 \tab 8 \tab 3 \tab 17 \tab 28 \tab 39 \tab 61 \tab 9 \tab 2 \tab 11 \tab 5 \tab 12 \tab 8 \tab 28 \tab 8 \tab 62 \tab 9 \tab 1 \tab 32 \tab 11 \tab 6 \tab 13 \tab 4 \tab 4 \tab 43 \tab 38 \tab 13 \tab 6 \tab 12 \tab 22 \tab 23 \tab 1 \tab 197 \tab 3 \tab - \tab - \tab 8 \tab 37 \tab 27 \tab 7 \tab 11 \tab 4 \tab 6 \tab - \tab 6 \tab 11 \tab 9 \tab 10 \tab 2 \tab 10 \tab 2 \tab - \tab 1 \tab 7 \tab 23 \tab 11 \tab 1 \tab 4 \tab 6 \tab 20 \tab 16 \tab 8 \tab 26 \tab 12 \tab 227 \tab 23 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 3 \tab 5 \tab 18 \tab 2 \tab - \tab 8 \tab 21 \tab 3 \tab 60 \tab 15 \tab 5 \tab 11 \tab 6 \tab 27 \tab 6 \tab 6 \tab 22 \tab 17 \tab 27 \tab 91 \tab 11 \tab 8 \tab 7 \tab 4 \tab 10 \tab 15 \tab 27 \tab 6 \tab 33 \tab 9 \tab 2 \tab 42 \tab 10 \tab 7 \tab 13 \tab 7 \tab 3 \tab 33 \tab 47 \tab 12 \tab 5 \tab 15 \tab 24 \tab 18 \tab 3 \tab 210 \tab 7 \tab 1 \tab 1 \tab 11 \tab 48 \tab 24 \tab 4 \tab 13 \tab 2 \tab 16 \tab 2 \tab 3 \tab 10 \tab 11 \tab 9 \tab - \tab 17 \tab 1 \tab 3 \tab 2 \tab 8 \tab 18 \tab 12 \tab 1 \tab 1 \tab 6 \tab 18 \tab 24 \tab 5 \tab 27 \tab 17 \tab 247 \tab 19 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1537 \cr
   1990 \tab 2 \tab 5 \tab 8 \tab - \tab 3 \tab 11 \tab 23 \tab 4 \tab 45 \tab 18 \tab 9 \tab 6 \tab 2 \tab 24 \tab 4 \tab 3 \tab 29 \tab 20 \tab 34 \tab 62 \tab 15 \tab 4 \tab 17 \tab 8 \tab 24 \tab 7 \tab 21 \tab 6 \tab 45 \tab 12 \tab - \tab 32 \tab 6 \tab 7 \tab 13 \tab 12 \tab 7 \tab 43 \tab 41 \tab 8 \tab 6 \tab 10 \tab 11 \tab 19 \tab 4 \tab 180 \tab 6 \tab 1 \tab 1 \tab 11 \tab 34 \tab 11 \tab 3 \tab 12 \tab 6 \tab 5 \tab 1 \tab 6 \tab 8 \tab 10 \tab 4 \tab 1 \tab 8 \tab 4 \tab 1 \tab 2 \tab 6 \tab 16 \tab 9 \tab 3 \tab 3 \tab 2 \tab 15 \tab 10 \tab 3 \tab 30 \tab 10 \tab 233 \tab 13 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1 \tab 5 \tab 8 \tab 3 \tab 3 \tab 21 \tab 23 \tab 3 \tab 56 \tab 11 \tab 8 \tab 13 \tab 6 \tab 26 \tab 8 \tab 9 \tab 31 \tab 18 \tab 27 \tab 77 \tab 9 \tab 6 \tab 11 \tab 4 \tab 13 \tab 8 \tab 20 \tab 5 \tab 54 \tab 15 \tab 2 \tab 39 \tab 10 \tab 8 \tab 11 \tab 1 \tab 10 \tab 38 \tab 39 \tab 15 \tab 5 \tab 12 \tab 32 \tab 32 \tab 3 \tab 187 \tab 1 \tab - \tab - \tab 8 \tab 33 \tab 24 \tab 5 \tab 9 \tab 5 \tab 6 \tab 1 \tab 5 \tab 12 \tab 10 \tab 4 \tab 1 \tab 10 \tab 1 \tab 3 \tab 6 \tab 2 \tab 24 \tab 9 \tab 8 \tab 1 \tab 3 \tab 8 \tab 19 \tab 6 \tab 28 \tab 20 \tab 266 \tab 20 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 5 \tab 6 \tab 2 \tab 2 \tab 20 \tab 32 \tab 5 \tab 74 \tab 21 \tab 6 \tab 17 \tab 5 \tab 15 \tab 8 \tab 6 \tab 32 \tab 18 \tab 32 \tab 71 \tab 14 \tab 8 \tab 15 \tab 8 \tab 18 \tab 14 \tab 27 \tab 13 \tab 45 \tab 12 \tab - \tab 35 \tab 8 \tab 10 \tab 9 \tab 2 \tab 5 \tab 43 \tab 38 \tab 8 \tab - \tab 14 \tab 16 \tab 25 \tab 3 \tab 205 \tab 5 \tab 1 \tab 2 \tab 18 \tab 38 \tab 28 \tab - \tab 13 \tab 2 \tab 8 \tab 2 \tab 17 \tab 17 \tab 11 \tab 16 \tab 1 \tab 22 \tab 3 \tab - \tab - \tab 5 \tab 20 \tab 12 \tab 4 \tab 3 \tab 12 \tab 18 \tab 23 \tab 7 \tab 25 \tab 22 \tab 244 \tab 20 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 4 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 14 \tab 14 \tab 2 \tab 9 \tab 49 \tab 55 \tab 5 \tab 92 \tab 34 \tab 16 \tab 21 \tab 4 \tab 49 \tab 19 \tab 6 \tab 67 \tab 35 \tab 56 \tab 154 \tab 12 \tab 10 \tab 27 \tab 26 \tab 43 \tab 21 \tab 57 \tab 16 \tab 72 \tab 22 \tab 1 \tab 75 \tab 12 \tab 14 \tab 16 \tab 6 \tab 13 \tab 75 \tab 74 \tab 19 \tab 5 \tab 24 \tab 35 \tab 42 \tab 5 \tab 318 \tab 7 \tab 3 \tab 8 \tab 26 \tab 60 \tab 37 \tab 7 \tab 16 \tab 5 \tab 10 \tab 3 \tab 21 \tab 19 \tab 15 \tab 17 \tab 8 \tab 33 \tab 11 \tab 1 \tab 7 \tab 18 \tab 53 \tab 26 \tab 15 \tab 2 \tab 25 \tab 46 \tab 46 \tab 12 \tab 58 \tab 22 \tab 551 \tab 31 \tab 3 \tab 3 \tab 2 \tab 1 \tab 1 \tab 3 \tab 4 \tab 2 \tab 3 \tab 2 \tab 5 \tab - \tab 2 \tab 1 \tab - \tab 2992 \cr
   1996 \tab 3 \tab 12 \tab 22 \tab 5 \tab 4 \tab 34 \tab 47 \tab 2 \tab 100 \tab 34 \tab 17 \tab 21 \tab 7 \tab 57 \tab 16 \tab 12 \tab 50 \tab 20 \tab 64 \tab 128 \tab 12 \tab 14 \tab 15 \tab 16 \tab 36 \tab 18 \tab 60 \tab 15 \tab 78 \tab 22 \tab - \tab 71 \tab 9 \tab 14 \tab 14 \tab 6 \tab 9 \tab 65 \tab 77 \tab 18 \tab 7 \tab 18 \tab 44 \tab 49 \tab 8 \tab 270 \tab 10 \tab 1 \tab 2 \tab 24 \tab 78 \tab 55 \tab 7 \tab 30 \tab 5 \tab 15 \tab 3 \tab 15 \tab 22 \tab 9 \tab 19 \tab 2 \tab 34 \tab 7 \tab 6 \tab 8 \tab 8 \tab 55 \tab 23 \tab 10 \tab 7 \tab 18 \tab 41 \tab 48 \tab 10 \tab 59 \tab 25 \tab 568 \tab 34 \tab 1 \tab 4 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 5 \tab - \tab 2 \tab 4 \tab - \tab 3 \tab 1 \tab - \tab 2904 \cr
   1998 \tab 2 \tab 8 \tab 15 \tab 3 \tab 6 \tab 32 \tab 30 \tab 3 \tab 88 \tab 36 \tab 16 \tab 18 \tab 8 \tab 35 \tab 20 \tab 8 \tab 57 \tab 43 \tab 58 \tab 113 \tab 16 \tab 7 \tab 22 \tab 10 \tab 25 \tab 29 \tab 60 \tab 20 \tab 80 \tab 26 \tab - \tab 74 \tab 11 \tab 15 \tab 15 \tab 4 \tab 14 \tab 66 \tab 70 \tab 24 \tab 11 \tab 19 \tab 48 \tab 30 \tab 6 \tab 255 \tab 5 \tab 4 \tab 6 \tab 32 \tab 74 \tab 45 \tab 7 \tab 29 \tab 11 \tab 15 \tab 3 \tab 13 \tab 28 \tab 12 \tab 25 \tab - \tab 24 \tab 7 \tab 4 \tab - \tab 9 \tab 48 \tab 22 \tab 11 \tab 5 \tab 22 \tab 32 \tab 47 \tab 9 \tab 65 \tab 28 \tab 569 \tab 33 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab 5 \tab 5 \tab 2 \tab 1 \tab 8 \tab 1 \tab 1 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 3 \tab 16 \tab 16 \tab 3 \tab 5 \tab 37 \tab 46 \tab 6 \tab 92 \tab 38 \tab 25 \tab 25 \tab 10 \tab 49 \tab 13 \tab 13 \tab 37 \tab 35 \tab 67 \tab 151 \tab 10 \tab 6 \tab 18 \tab 7 \tab 31 \tab 19 \tab 38 \tab 16 \tab 68 \tab 16 \tab 4 \tab 65 \tab 16 \tab 20 \tab 16 \tab 8 \tab 7 \tab 64 \tab 81 \tab 13 \tab 8 \tab 24 \tab 33 \tab 43 \tab 5 \tab 274 \tab 16 \tab 3 \tab 1 \tab 24 \tab 61 \tab 35 \tab 7 \tab 19 \tab 6 \tab 11 \tab 8 \tab 8 \tab 41 \tab 16 \tab 12 \tab 3 \tab 35 \tab 12 \tab 1 \tab 2 \tab 9 \tab 38 \tab 21 \tab 18 \tab 2 \tab 17 \tab 29 \tab 42 \tab 7 \tab 60 \tab 30 \tab 576 \tab 27 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 6 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab - \tab 2817 \cr
   2002 \tab 1 \tab 11 \tab 9 \tab 1 \tab 6 \tab 28 \tab 42 \tab 4 \tab 91 \tab 34 \tab 16 \tab 22 \tab 7 \tab 46 \tab 17 \tab 11 \tab 38 \tab 32 \tab 79 \tab 132 \tab 16 \tab 9 \tab 21 \tab 8 \tab 30 \tab 17 \tab 38 \tab 12 \tab 67 \tab 24 \tab 2 \tab 64 \tab 10 \tab 13 \tab 17 \tab 10 \tab 10 \tab 67 \tab 75 \tab 23 \tab 12 \tab 17 \tab 50 \tab 35 \tab 3 \tab 234 \tab 14 \tab 2 \tab 7 \tab 13 \tab 64 \tab 40 \tab 4 \tab 26 \tab 4 \tab 11 \tab 7 \tab 16 \tab 25 \tab 26 \tab 27 \tab 4 \tab 28 \tab 9 \tab 1 \tab 5 \tab 11 \tab 52 \tab 17 \tab 11 \tab 3 \tab 13 \tab 28 \tab 48 \tab 8 \tab 49 \tab 24 \tab 589 \tab 36 \tab 1 \tab 6 \tab 2 \tab 4 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 8 \tab - \tab 1 \tab 1 \tab - \tab 2765 \cr
   2004 \tab 1 \tab 12 \tab 9 \tab - \tab 1 \tab 32 \tab 34 \tab 4 \tab 117 \tab 37 \tab 16 \tab 17 \tab 6 \tab 64 \tab 10 \tab 11 \tab 52 \tab 40 \tab 71 \tab 127 \tab 12 \tab 14 \tab 23 \tab 10 \tab 29 \tab 19 \tab 51 \tab 19 \tab 64 \tab 13 \tab - \tab 71 \tab 11 \tab 12 \tab 17 \tab 11 \tab 5 \tab 70 \tab 79 \tab 20 \tab 8 \tab 25 \tab 36 \tab 30 \tab 8 \tab 197 \tab 10 \tab 1 \tab 4 \tab 19 \tab 76 \tab 30 \tab 8 \tab 22 \tab 5 \tab 8 \tab 6 \tab 16 \tab 29 \tab 24 \tab 19 \tab 3 \tab 27 \tab 10 \tab 3 \tab 7 \tab 4 \tab 55 \tab 18 \tab 11 \tab 7 \tab 22 \tab 28 \tab 47 \tab 12 \tab 63 \tab 35 \tab 603 \tab 30 \tab 1 \tab 4 \tab - \tab 2 \tab 1 \tab 4 \tab 6 \tab 6 \tab - \tab - \tab 5 \tab 1 \tab 5 \tab - \tab - \tab 2812 \cr
   2006 \tab 3 \tab 11 \tab 16 \tab 2 \tab 9 \tab 35 \tab 40 \tab 4 \tab 118 \tab 50 \tab 14 \tab 32 \tab 13 \tab 63 \tab 15 \tab 13 \tab 57 \tab 39 \tab 72 \tab 133 \tab 11 \tab 6 \tab 25 \tab 14 \tab 34 \tab 23 \tab 39 \tab 14 \tab 73 \tab 19 \tab 2 \tab 69 \tab 5 \tab 16 \tab 21 \tab 6 \tab 11 \tab 63 \tab 86 \tab 21 \tab 12 \tab 29 \tab 56 \tab 39 \tab 5 \tab 218 \tab 10 \tab 1 \tab 10 \tab 34 \tab 52 \tab 54 \tab 9 \tab 33 \tab 10 \tab 8 \tab 5 \tab 7 \tab 27 \tab 35 \tab 16 \tab 2 \tab 30 \tab 5 \tab 2 \tab 3 \tab 16 \tab 62 \tab 28 \tab 10 \tab 6 \tab 23 \tab 35 \tab 39 \tab 12 \tab 56 \tab 38 \tab 2107 \tab 40 \tab - \tab 5 \tab 2 \tab 1 \tab - \tab 3 \tab 5 \tab 3 \tab 1 \tab 3 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 4510 \cr
   2008 \tab - \tab 11 \tab 13 \tab - \tab 2 \tab 27 \tab 26 \tab 7 \tab 57 \tab 28 \tab 11 \tab 12 \tab 3 \tab 41 \tab 10 \tab 8 \tab 28 \tab 28 \tab 63 \tab 84 \tab 14 \tab 6 \tab 16 \tab 17 \tab 19 \tab 16 \tab 43 \tab 11 \tab 29 \tab 15 \tab 1 \tab 58 \tab 8 \tab 9 \tab 8 \tab 4 \tab 8 \tab 40 \tab 66 \tab 14 \tab 7 \tab 9 \tab 42 \tab 20 \tab 6 \tab 144 \tab 6 \tab 3 \tab 4 \tab 29 \tab 42 \tab 32 \tab 12 \tab 15 \tab 9 \tab 14 \tab 3 \tab 13 \tab 28 \tab 29 \tab 13 \tab 3 \tab 20 \tab 3 \tab 6 \tab 4 \tab 6 \tab 40 \tab 11 \tab 4 \tab 6 \tab 11 \tab 22 \tab 23 \tab 5 \tab 50 \tab 19 \tab 398 \tab 30 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 6 \tab 2 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 2023 \cr
   2010 \tab - \tab 13 \tab 12 \tab 2 \tab 1 \tab 29 \tab 30 \tab 6 \tab 68 \tab 19 \tab 12 \tab 19 \tab 10 \tab 51 \tab 12 \tab 5 \tab 29 \tab 27 \tab 35 \tab 98 \tab 8 \tab 8 \tab 18 \tab 5 \tab 16 \tab 15 \tab 27 \tab 17 \tab 52 \tab 20 \tab 1 \tab 57 \tab 9 \tab 15 \tab 17 \tab 7 \tab 8 \tab 44 \tab 64 \tab 12 \tab 5 \tab 26 \tab 22 \tab 28 \tab 11 \tab 135 \tab 8 \tab 2 \tab 3 \tab 13 \tab 47 \tab 28 \tab 8 \tab 24 \tab 4 \tab 7 \tab 8 \tab 9 \tab 16 \tab 19 \tab 15 \tab 2 \tab 17 \tab 7 \tab 3 \tab 4 \tab 5 \tab 37 \tab 8 \tab 11 \tab 6 \tab 11 \tab 23 \tab 29 \tab 4 \tab 55 \tab 28 \tab 402 \tab 33 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 2 \tab 7 \tab - \tab 1 \tab 1 \tab - \tab 2044 \cr
   2012 \tab 3 \tab 4 \tab 11 \tab 1 \tab 3 \tab 31 \tab 30 \tab 3 \tab 77 \tab 20 \tab 13 \tab 9 \tab 2 \tab 60 \tab 15 \tab 7 \tab 31 \tab 25 \tab 66 \tab 81 \tab 8 \tab 6 \tab 12 \tab 3 \tab 26 \tab 11 \tab 30 \tab 5 \tab 47 \tab 9 \tab 1 \tab 64 \tab 9 \tab 8 \tab 2 \tab 8 \tab 5 \tab 41 \tab 53 \tab 16 \tab 1 \tab 22 \tab 24 \tab 26 \tab 2 \tab 123 \tab 9 \tab 1 \tab 5 \tab 19 \tab 29 \tab 20 \tab 8 \tab 16 \tab 5 \tab 5 \tab 2 \tab 11 \tab 24 \tab - \tab 15 \tab 5 \tab 27 \tab 4 \tab 5 \tab 2 \tab 10 \tab 73 \tab 16 \tab 11 \tab 5 \tab 10 \tab 19 \tab 30 \tab 7 \tab 50 \tab 49 \tab 385 \tab 16 \tab 2 \tab 6 \tab 1 \tab 1 \tab - \tab 3 \tab 4 \tab 1 \tab - \tab - \tab 8 \tab - \tab - \tab 1 \tab - \tab 1974 \cr
   2014 \tab 1 \tab 13 \tab 10 \tab 3 \tab - \tab 33 \tab 44 \tab 2 \tab 97 \tab 24 \tab 11 \tab 9 \tab 5 \tab 61 \tab 16 \tab 4 \tab 51 \tab 32 \tab 67 \tab 112 \tab 14 \tab 7 \tab 19 \tab 10 \tab 26 \tab 16 \tab 30 \tab 13 \tab 71 \tab 22 \tab 1 \tab 47 \tab 25 \tab 18 \tab 9 \tab 6 \tab 5 \tab 62 \tab 83 \tab 19 \tab 10 \tab 17 \tab 37 \tab 48 \tab 5 \tab 156 \tab 11 \tab 1 \tab 3 \tab 13 \tab 60 \tab 34 \tab 8 \tab 30 \tab 7 \tab 6 \tab 6 \tab 7 \tab 32 \tab 50 \tab 24 \tab 1 \tab 19 \tab 10 \tab 4 \tab 3 \tab 5 \tab 37 \tab 18 \tab 15 \tab 4 \tab 19 \tab 29 \tab 42 \tab 5 \tab 56 \tab 23 \tab 524 \tab 32 \tab - \tab 5 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 8 \tab - \tab 3 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 16 \tab 17 \tab 2 \tab 1 \tab 58 \tab 36 \tab 4 \tab 121 \tab 23 \tab 18 \tab 22 \tab 9 \tab 82 \tab 19 \tab 16 \tab 48 \tab 36 \tab 75 \tab 138 \tab 12 \tab 8 \tab 24 \tab 13 \tab 36 \tab 9 \tab 45 \tab 13 \tab 59 \tab 20 \tab 5 \tab 71 \tab 21 \tab 14 \tab 13 \tab 6 \tab 8 \tab 58 \tab 95 \tab 12 \tab 12 \tab 14 \tab 38 \tab 22 \tab 8 \tab 144 \tab 17 \tab 1 \tab 7 \tab 17 \tab 55 \tab 30 \tab 12 \tab 42 \tab 9 \tab 11 \tab 12 \tab 18 \tab 25 \tab 26 \tab 34 \tab 2 \tab 21 \tab 13 \tab 1 \tab 7 \tab 24 \tab 42 \tab 16 \tab 14 \tab 9 \tab 15 \tab 30 \tab 53 \tab 4 \tab 63 \tab 28 \tab 625 \tab 28 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab 5 \tab 7 \tab 1 \tab 3 \tab 7 \tab - \tab 1 \tab 1 \tab - \tab 2867 \cr
   2018 \tab - \tab 8 \tab 12 \tab 1 \tab 7 \tab 36 \tab 37 \tab 14 \tab 90 \tab 22 \tab 14 \tab 20 \tab 10 \tab 56 \tab 14 \tab 7 \tab 34 \tab 25 \tab 54 \tab 106 \tab 12 \tab 13 \tab 18 \tab 6 \tab 29 \tab 8 \tab 26 \tab 14 \tab 39 \tab 8 \tab 1 \tab 39 \tab 15 \tab 14 \tab 10 \tab 6 \tab 8 \tab 49 \tab 70 \tab 21 \tab 43 \tab 22 \tab 36 \tab 26 \tab 10 \tab 114 \tab 14 \tab 4 \tab 8 \tab 20 \tab 37 \tab 63 \tab 4 \tab 28 \tab 5 \tab 14 \tab 6 \tab 10 \tab 26 \tab 28 \tab 18 \tab 1 \tab 32 \tab 9 \tab 4 \tab 6 \tab 12 \tab 31 \tab 16 \tab 13 \tab 8 \tab 11 \tab 39 \tab 32 \tab 3 \tab 35 \tab 34 \tab 498 \tab 23 \tab 2 \tab 4 \tab - \tab 1 \tab 2 \tab 4 \tab 7 \tab 2 \tab 3 \tab - \tab 4 \tab - \tab 3 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 45 \tab 271 \tab 350 \tab 68 \tab 105 \tab 772 \tab 973 \tab 142 \tab 2458 \tab 790 \tab 355 \tab 506 \tab 181 \tab 1160 \tab 404 \tab 227 \tab 1055 \tab 796 \tab 1361 \tab 2876 \tab 310 \tab 215 \tab 450 \tab 290 \tab 673 \tab 487 \tab 983 \tab 363 \tab 1546 \tab 421 \tab 36 \tab 1639 \tab 302 \tab 330 \tab 421 \tab 176 \tab 229 \tab 1668 \tab 1724 \tab 415 \tab 200 \tab 550 \tab 878 \tab 804 \tab 138 \tab 6953 \tab 216 \tab 41 \tab 105 \tab 515 \tab 1446 \tab 902 \tab 165 \tab 534 \tab 145 \tab 275 \tab 108 \tab 268 \tab 516 \tab 460 \tab 430 \tab 71 \tab 558 \tab 170 \tab 67 \tab 80 \tab 258 \tab 979 \tab 433 \tab 205 \tab 103 \tab 343 \tab 638 \tab 804 \tab 170 \tab 1142 \tab 619 \tab 17254 \tab 655 \tab 28 \tab 81 \tab 32 \tab 32 \tab 12 \tab 73 \tab 73 \tab 65 \tab 21 \tab 24 \tab 105 \tab 5 \tab 26 \tab 10 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
