% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_06.R
\name{papres16}
\alias{papres16}
\title{papres16}
\source{
General Social Survey https://gss.norc.org
}
\description{
Father's occupational prestige score (1970)
}
\details{
Question 8a. What kind of work did your father \link{FATHER SUBSTITUTE} normally do while you were growing up? That is, what was his job called? b. What did he actually do in that job? Tell me, what were some of his main duties?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 12 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 60 \tab 61 \tab 62 \tab 63 \tab 67 \tab 68 \tab 69 \tab 70 \tab 71 \tab 72 \tab 74 \tab 76 \tab 78 \tab 82 \tab dk, na, iap \tab 14 \tab 53 \tab 15 \tab 9 \tab 66 \tab 65 \tab not available in this year \tab iap \tab Total \cr
   1972 \tab 4 \tab 14 \tab 63 \tab 44 \tab 11 \tab 4 \tab 1 \tab 15 \tab 11 \tab 3 \tab 7 \tab 46 \tab 4 \tab 11 \tab 26 \tab 18 \tab 11 \tab 82 \tab 31 \tab 36 \tab 17 \tab 49 \tab 33 \tab 17 \tab 18 \tab 44 \tab 342 \tab 9 \tab 7 \tab 14 \tab 37 \tab 3 \tab 23 \tab 31 \tab 11 \tab 138 \tab 12 \tab 2 \tab 3 \tab 2 \tab 2 \tab 11 \tab 4 \tab 7 \tab 10 \tab 7 \tab 5 \tab 3 \tab 2 \tab 17 \tab 1 \tab 4 \tab 5 \tab 3 \tab 4 \tab 1 \tab 7 \tab 266 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1 \tab 5 \tab 62 \tab 37 \tab 6 \tab 7 \tab - \tab 17 \tab 11 \tab 6 \tab 6 \tab 62 \tab 12 \tab 13 \tab 18 \tab 16 \tab 8 \tab 88 \tab 28 \tab 20 \tab 16 \tab 33 \tab 20 \tab 9 \tab 26 \tab 63 \tab 273 \tab 10 \tab 5 \tab 55 \tab 56 \tab 1 \tab 22 \tab 37 \tab 15 \tab 134 \tab 12 \tab 3 \tab 4 \tab 1 \tab 4 \tab 15 \tab 4 \tab 8 \tab 11 \tab 8 \tab 2 \tab 2 \tab 5 \tab 19 \tab - \tab 4 \tab 5 \tab 2 \tab 3 \tab 3 \tab 5 \tab 183 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 6 \tab 7 \tab 52 \tab 46 \tab 1 \tab 2 \tab 1 \tab 9 \tab 11 \tab 4 \tab 4 \tab 52 \tab 13 \tab 26 \tab 37 \tab 8 \tab 14 \tab 65 \tab 40 \tab 14 \tab 21 \tab 31 \tab 41 \tab 12 \tab 31 \tab 65 \tab 291 \tab 9 \tab 9 \tab 5 \tab 39 \tab 4 \tab 26 \tab 28 \tab 12 \tab 142 \tab 15 \tab 2 \tab 1 \tab 3 \tab 4 \tab 15 \tab 3 \tab 7 \tab 7 \tab 2 \tab 1 \tab 3 \tab 4 \tab 16 \tab - \tab 2 \tab 2 \tab 6 \tab 2 \tab 4 \tab 3 \tab 201 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 2 \tab 10 \tab 62 \tab 20 \tab 7 \tab 5 \tab 1 \tab 15 \tab 7 \tab 4 \tab 7 \tab 51 \tab 18 \tab 37 \tab 15 \tab 15 \tab 8 \tab 78 \tab 36 \tab 34 \tab 18 \tab 34 \tab 34 \tab 14 \tab 24 \tab 65 \tab 293 \tab 9 \tab 7 \tab 20 \tab 41 \tab 3 \tab 27 \tab 44 \tab 10 \tab 120 \tab 14 \tab 1 \tab 4 \tab 1 \tab 5 \tab 9 \tab 3 \tab 13 \tab 13 \tab 10 \tab 2 \tab 1 \tab 3 \tab 14 \tab - \tab 3 \tab 7 \tab 2 \tab 1 \tab 3 \tab 7 \tab 176 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 3 \tab 16 \tab 80 \tab 26 \tab 5 \tab 4 \tab - \tab 9 \tab 14 \tab 3 \tab 3 \tab 47 \tab 17 \tab 24 \tab 26 \tab 18 \tab 12 \tab 84 \tab 27 \tab 32 \tab 25 \tab 38 \tab 29 \tab 8 \tab 25 \tab 49 \tab 304 \tab 6 \tab 6 \tab 12 \tab 53 \tab 1 \tab 32 \tab 32 \tab 10 \tab 125 \tab 11 \tab - \tab 2 \tab - \tab 5 \tab 11 \tab 1 \tab 5 \tab - \tab 2 \tab 5 \tab - \tab 3 \tab 13 \tab - \tab 1 \tab 3 \tab 2 \tab 5 \tab 3 \tab 10 \tab 209 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 3 \tab 14 \tab 57 \tab 46 \tab 5 \tab 4 \tab 1 \tab 9 \tab 12 \tab 3 \tab 4 \tab 75 \tab 15 \tab 17 \tab 25 \tab 21 \tab 8 \tab 86 \tab 44 \tab 30 \tab 15 \tab 35 \tab 37 \tab 11 \tab 31 \tab 52 \tab 268 \tab 12 \tab 8 \tab 4 \tab 36 \tab 3 \tab 38 \tab 33 \tab 13 \tab 121 \tab 14 \tab 1 \tab 2 \tab 2 \tab 10 \tab 9 \tab 1 \tab 14 \tab 7 \tab 4 \tab 1 \tab 2 \tab 5 \tab 13 \tab 2 \tab 1 \tab 8 \tab 3 \tab 10 \tab 3 \tab 4 \tab 215 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 3 \tab 12 \tab 43 \tab 32 \tab 8 \tab 4 \tab - \tab 6 \tab 6 \tab 7 \tab 6 \tab 75 \tab 12 \tab 27 \tab 26 \tab 17 \tab 8 \tab 78 \tab 44 \tab 46 \tab 14 \tab 37 \tab 27 \tab 10 \tab 20 \tab 72 \tab 255 \tab 24 \tab 2 \tab 13 \tab 33 \tab 3 \tab 45 \tab 39 \tab 18 \tab 130 \tab 12 \tab 2 \tab 2 \tab 4 \tab 10 \tab 12 \tab 3 \tab 8 \tab 6 \tab 2 \tab 5 \tab 1 \tab 5 \tab 18 \tab - \tab 1 \tab 5 \tab - \tab 8 \tab 1 \tab 7 \tab 214 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1 \tab 12 \tab 54 \tab 18 \tab 6 \tab 4 \tab - \tab 12 \tab 5 \tab 2 \tab 1 \tab 64 \tab 16 \tab 17 \tab 19 \tab 17 \tab 5 \tab 63 \tab 29 \tab 28 \tab 22 \tab 35 \tab 25 \tab 15 \tab 24 \tab 62 \tab 256 \tab 7 \tab 5 \tab 16 \tab 33 \tab 1 \tab 41 \tab 36 \tab 13 \tab 144 \tab 9 \tab 3 \tab 6 \tab - \tab 7 \tab 12 \tab 5 \tab 8 \tab 10 \tab 3 \tab 4 \tab 2 \tab 2 \tab 13 \tab 1 \tab 3 \tab 3 \tab 1 \tab 11 \tab 9 \tab 8 \tab 238 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 10 \tab 17 \tab 86 \tab 44 \tab 6 \tab 5 \tab - \tab 23 \tab 14 \tab 2 \tab 2 \tab 72 \tab 16 \tab 27 \tab 18 \tab 17 \tab 8 \tab 141 \tab 40 \tab 55 \tab 15 \tab 49 \tab 38 \tab 12 \tab 26 \tab 63 \tab 294 \tab 17 \tab 7 \tab 18 \tab 33 \tab 1 \tab 39 \tab 26 \tab 12 \tab 124 \tab 13 \tab 2 \tab 9 \tab 1 \tab 7 \tab 5 \tab - \tab 9 \tab 9 \tab 5 \tab 8 \tab 1 \tab 6 \tab 17 \tab 1 \tab 4 \tab 6 \tab 3 \tab 5 \tab 2 \tab 5 \tab 357 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 24 \tab 63 \tab 23 \tab 4 \tab 5 \tab - \tab 11 \tab 18 \tab 3 \tab 10 \tab 42 \tab 15 \tab 25 \tab 32 \tab 16 \tab 12 \tab 83 \tab 32 \tab 45 \tab 29 \tab 40 \tab 35 \tab 16 \tab 33 \tab 68 \tab 213 \tab 14 \tab 5 \tab 9 \tab 54 \tab 2 \tab 29 \tab 38 \tab 17 \tab 152 \tab 10 \tab 3 \tab 2 \tab - \tab 4 \tab 15 \tab 3 \tab 11 \tab 7 \tab 13 \tab 1 \tab 4 \tab 7 \tab 21 \tab 1 \tab 6 \tab 12 \tab 6 \tab 10 \tab 13 \tab 8 \tab 222 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 4 \tab 9 \tab 82 \tab 25 \tab 3 \tab 3 \tab - \tab 8 \tab 10 \tab 4 \tab 2 \tab 43 \tab 7 \tab 14 \tab 31 \tab 17 \tab 9 \tab 80 \tab 32 \tab 36 \tab 25 \tab 40 \tab 12 \tab 9 \tab 20 \tab 34 \tab 264 \tab 8 \tab 2 \tab 17 \tab 28 \tab 1 \tab 47 \tab 24 \tab 17 \tab 171 \tab 13 \tab 2 \tab 3 \tab 1 \tab 9 \tab 7 \tab 2 \tab 10 \tab 10 \tab 2 \tab 3 \tab 4 \tab 4 \tab 16 \tab - \tab 3 \tab 3 \tab 1 \tab 5 \tab 5 \tab 3 \tab 224 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 3 \tab 12 \tab 56 \tab 27 \tab 4 \tab 7 \tab 2 \tab 6 \tab 9 \tab 3 \tab 3 \tab 23 \tab 10 \tab 14 \tab 12 \tab 11 \tab 6 \tab 93 \tab 40 \tab 40 \tab 19 \tab 46 \tab 39 \tab 7 \tab 22 \tab 53 \tab 243 \tab 17 \tab 10 \tab 15 \tab 64 \tab 1 \tab 39 \tab 38 \tab 18 \tab 159 \tab 18 \tab 3 \tab 2 \tab - \tab 6 \tab 11 \tab 2 \tab 6 \tab 14 \tab 11 \tab 5 \tab 10 \tab 6 \tab 25 \tab - \tab 3 \tab 10 \tab 4 \tab 8 \tab 5 \tab 9 \tab 203 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 3 \tab 11 \tab 56 \tab 25 \tab 5 \tab 5 \tab - \tab 10 \tab 10 \tab - \tab 2 \tab 18 \tab 8 \tab 2 \tab 25 \tab 16 \tab 17 \tab 82 \tab 30 \tab 61 \tab 9 \tab 19 \tab 20 \tab 7 \tab 22 \tab 55 \tab 234 \tab 12 \tab 6 \tab 13 \tab 48 \tab 2 \tab 35 \tab 52 \tab 16 \tab 165 \tab 14 \tab - \tab 4 \tab 4 \tab 8 \tab 4 \tab 4 \tab 5 \tab 11 \tab 6 \tab 2 \tab 7 \tab 4 \tab 18 \tab 1 \tab 5 \tab 3 \tab 4 \tab 9 \tab 7 \tab 5 \tab 244 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1 \tab 14 \tab 99 \tab 46 \tab 7 \tab 6 \tab - \tab 14 \tab 10 \tab - \tab 4 \tab 50 \tab 16 \tab 26 \tab 18 \tab 18 \tab 14 \tab 141 \tab 17 \tab 44 \tab 21 \tab 43 \tab 28 \tab 4 \tab 17 \tab 70 \tab 300 \tab 12 \tab 5 \tab 14 \tab 42 \tab 1 \tab 51 \tab 27 \tab 14 \tab 146 \tab 11 \tab - \tab 2 \tab 1 \tab 6 \tab 20 \tab 3 \tab 5 \tab 13 \tab 7 \tab 10 \tab 7 \tab 7 \tab 24 \tab - \tab 5 \tab 5 \tab 2 \tab 9 \tab 9 \tab 11 \tab 315 \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 7 \tab 8 \tab 56 \tab 38 \tab 4 \tab 2 \tab - \tab 17 \tab 15 \tab - \tab 6 \tab 31 \tab 10 \tab 11 \tab 23 \tab 10 \tab 13 \tab 84 \tab 38 \tab 53 \tab 8 \tab 25 \tab 36 \tab 9 \tab 16 \tab 49 \tab 212 \tab 14 \tab 2 \tab 8 \tab 31 \tab 1 \tab 36 \tab 48 \tab 14 \tab 167 \tab 8 \tab 2 \tab 1 \tab 2 \tab 9 \tab 15 \tab 1 \tab 9 \tab 19 \tab 9 \tab 7 \tab 12 \tab 5 \tab 21 \tab - \tab 9 \tab 2 \tab 4 \tab 5 \tab 8 \tab 7 \tab 222 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1 \tab 9 \tab 63 \tab 22 \tab 5 \tab 3 \tab - \tab 10 \tab 12 \tab 7 \tab 5 \tab 37 \tab 16 \tab 14 \tab 14 \tab 7 \tab 5 \tab 104 \tab 34 \tab 33 \tab 16 \tab 36 \tab 39 \tab 9 \tab 12 \tab 43 \tab 222 \tab 18 \tab 7 \tab 19 \tab 38 \tab 3 \tab 28 \tab 43 \tab 16 \tab 176 \tab 17 \tab 4 \tab 5 \tab 1 \tab 8 \tab 11 \tab 3 \tab 16 \tab 11 \tab 7 \tab 8 \tab 5 \tab 9 \tab 22 \tab 3 \tab 7 \tab 7 \tab - \tab 6 \tab 12 \tab 7 \tab 242 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 11 \tab 66 \tab 21 \tab 3 \tab 2 \tab 1 \tab 5 \tab 9 \tab 3 \tab 2 \tab 23 \tab 15 \tab 15 \tab 19 \tab 10 \tab 8 \tab 77 \tab 29 \tab 41 \tab 10 \tab 27 \tab 27 \tab 5 \tab 23 \tab 51 \tab 181 \tab 17 \tab 8 \tab 10 \tab 40 \tab 3 \tab 28 \tab 30 \tab 8 \tab 162 \tab 15 \tab 1 \tab 6 \tab 1 \tab 10 \tab 11 \tab 2 \tab 8 \tab 14 \tab 11 \tab 5 \tab 3 \tab 8 \tab 22 \tab 1 \tab 8 \tab 5 \tab 2 \tab 1 \tab 6 \tab 6 \tab 234 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 52 \tab 205 \tab 1100 \tab 540 \tab 90 \tab 72 \tab 7 \tab 196 \tab 184 \tab 54 \tab 74 \tab 811 \tab 220 \tab 320 \tab 384 \tab 252 \tab 166 \tab 1509 \tab 571 \tab 648 \tab 300 \tab 617 \tab 520 \tab 174 \tab 390 \tab 958 \tab 4445 \tab 215 \tab 101 \tab 262 \tab 706 \tab 34 \tab 586 \tab 606 \tab 234 \tab 2476 \tab 218 \tab 31 \tab 58 \tab 24 \tab 114 \tab 193 \tab 44 \tab 149 \tab 172 \tab 109 \tab 74 \tab 67 \tab 85 \tab 309 \tab 11 \tab 69 \tab 91 \tab 45 \tab 102 \tab 94 \tab 112 \tab 3965 \tab 15 \tab 12 \tab 16 \tab 2 \tab 4 \tab 1 \tab 42581 \tab 3544 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
