% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_07.R
\name{papres80}
\alias{papres80}
\title{Father's occupational prestige score (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
papres80
}
\details{
Question 9. Father's occupation, prestige, and industry.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 68 \tab 69 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 86 \tab dk, na, iap \tab 19 \tab 55 \tab 56 \tab 67 \tab 17 \tab 20 \tab iap \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 2 \tab 19 \tab 33 \tab 34 \tab 1 \tab 6 \tab 12 \tab 16 \tab 13 \tab 57 \tab 20 \tab 30 \tab 26 \tab 41 \tab 44 \tab 49 \tab 14 \tab 7 \tab 46 \tab 218 \tab 13 \tab 37 \tab 16 \tab 46 \tab 24 \tab 1 \tab 56 \tab 7 \tab 33 \tab 17 \tab 143 \tab 4 \tab 5 \tab 22 \tab 1 \tab 1 \tab 8 \tab 18 \tab 1 \tab 1 \tab 4 \tab 26 \tab 10 \tab 8 \tab 2 \tab 17 \tab 3 \tab 5 \tab 6 \tab 14 \tab 10 \tab 5 \tab 7 \tab 222 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 2 \tab 13 \tab 32 \tab 49 \tab 5 \tab 2 \tab 10 \tab 14 \tab 7 \tab 66 \tab 22 \tab 25 \tab 30 \tab 33 \tab 45 \tab 53 \tab 9 \tab 6 \tab 41 \tab 244 \tab 14 \tab 44 \tab 20 \tab 44 \tab 28 \tab 6 \tab 62 \tab 10 \tab 36 \tab 15 \tab 120 \tab 6 \tab 10 \tab 15 \tab 1 \tab 1 \tab 10 \tab 20 \tab 6 \tab 5 \tab 1 \tab 27 \tab 9 \tab 8 \tab 1 \tab 23 \tab 1 \tab 2 \tab 3 \tab 7 \tab 15 \tab 6 \tab 6 \tab 242 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 2 \tab 6 \tab 29 \tab 39 \tab 2 \tab 3 \tab 7 \tab 7 \tab 13 \tab 65 \tab 23 \tab 23 \tab 17 \tab 27 \tab 43 \tab 60 \tab 5 \tab 5 \tab 38 \tab 190 \tab 11 \tab 56 \tab 23 \tab 49 \tab 27 \tab 4 \tab 55 \tab 7 \tab 38 \tab 16 \tab 109 \tab 5 \tab 10 \tab 2 \tab 1 \tab 2 \tab 5 \tab 11 \tab 3 \tab 5 \tab 2 \tab 25 \tab 8 \tab 6 \tab - \tab 17 \tab - \tab 1 \tab 7 \tab 5 \tab 8 \tab 2 \tab 5 \tab 235 \tab 4 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 2 \tab 21 \tab 17 \tab 32 \tab 11 \tab 4 \tab 13 \tab 8 \tab 10 \tab 68 \tab 15 \tab 23 \tab 12 \tab 40 \tab 48 \tab 52 \tab 4 \tab 10 \tab 40 \tab 195 \tab 8 \tab 38 \tab 22 \tab 59 \tab 28 \tab 7 \tab 54 \tab 37 \tab 60 \tab 20 \tab 111 \tab 5 \tab 4 \tab 19 \tab 1 \tab 1 \tab 9 \tab 7 \tab 8 \tab 5 \tab 3 \tab 30 \tab 13 \tab 9 \tab 2 \tab 15 \tab 1 \tab 5 \tab 9 \tab 7 \tab 3 \tab 4 \tab 10 \tab 266 \tab 5 \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 1517 \cr
   1993 \tab - \tab 5 \tab 20 \tab 21 \tab 47 \tab 2 \tab 2 \tab 18 \tab 11 \tab 17 \tab 62 \tab 15 \tab 24 \tab 16 \tab 32 \tab 31 \tab 45 \tab 6 \tab 9 \tab 49 \tab 260 \tab 9 \tab 52 \tab 28 \tab 80 \tab 27 \tab 4 \tab 64 \tab 9 \tab 51 \tab 26 \tab 117 \tab 2 \tab 15 \tab 13 \tab 1 \tab 2 \tab 4 \tab 21 \tab 7 \tab 4 \tab 2 \tab 23 \tab 12 \tab 10 \tab 2 \tab 24 \tab 3 \tab 10 \tab 7 \tab 7 \tab 9 \tab 10 \tab 13 \tab 239 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1606 \cr
   1994 \tab - \tab 6 \tab 42 \tab 30 \tab 92 \tab 5 \tab 8 \tab 24 \tab 21 \tab 15 \tab 143 \tab 41 \tab 33 \tab 21 \tab 71 \tab 84 \tab 108 \tab 26 \tab 21 \tab 64 \tab 397 \tab 33 \tab 99 \tab 30 \tab 125 \tab 47 \tab 12 \tab 101 \tab 17 \tab 83 \tab 42 \tab 208 \tab 5 \tab 25 \tab 26 \tab 1 \tab 2 \tab 15 \tab 25 \tab 16 \tab 7 \tab 5 \tab 55 \tab 28 \tab 14 \tab 6 \tab 35 \tab - \tab 16 \tab 13 \tab 25 \tab 18 \tab 24 \tab 12 \tab 561 \tab 4 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab 2992 \cr
   1996 \tab - \tab 2 \tab 34 \tab 37 \tab 81 \tab 6 \tab 6 \tab 17 \tab 20 \tab 14 \tab 138 \tab 41 \tab 46 \tab 37 \tab 65 \tab 75 \tab 114 \tab 14 \tab 19 \tab 71 \tab 308 \tab 29 \tab 80 \tab 29 \tab 97 \tab 45 \tab 15 \tab 111 \tab 31 \tab 47 \tab 45 \tab 259 \tab 9 \tab 19 \tab 35 \tab 2 \tab 5 \tab 19 \tab 22 \tab 26 \tab 2 \tab 9 \tab 67 \tab 26 \tab 21 \tab 5 \tab 39 \tab 2 \tab 5 \tab 12 \tab 18 \tab 13 \tab 16 \tab 15 \tab 578 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 7 \tab 40 \tab 35 \tab 75 \tab 9 \tab 6 \tab 9 \tab 15 \tab 18 \tab 111 \tab 28 \tab 44 \tab 46 \tab 55 \tab 97 \tab 117 \tab 4 \tab 15 \tab 98 \tab 312 \tab 13 \tab 62 \tab 26 \tab 98 \tab 56 \tab 8 \tab 115 \tab 13 \tab 71 \tab 35 \tab 230 \tab 7 \tab 19 \tab 25 \tab 6 \tab 2 \tab 16 \tab 34 \tab 21 \tab 8 \tab 3 \tab 57 \tab 32 \tab 23 \tab 7 \tab 41 \tab 2 \tab 7 \tab 6 \tab 12 \tab 22 \tab 19 \tab 16 \tab 571 \tab 1 \tab 1 \tab 2 \tab - \tab 4 \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 2 \tab 36 \tab 41 \tab 61 \tab 5 \tab 6 \tab 14 \tab 26 \tab 14 \tab 171 \tab 32 \tab 27 \tab 55 \tab 70 \tab 79 \tab 99 \tab 11 \tab 8 \tab 78 \tab 332 \tab 14 \tab 73 \tab 34 \tab 90 \tab 49 \tab 7 \tab 103 \tab 9 \tab 76 \tab 34 \tab 209 \tab 5 \tab 30 \tab 15 \tab 1 \tab 1 \tab 15 \tab 25 \tab 21 \tab 8 \tab 3 \tab 62 \tab 24 \tab 19 \tab 4 \tab 35 \tab 3 \tab 3 \tab 13 \tab 13 \tab 18 \tab 15 \tab 13 \tab 600 \tab 4 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2817 \cr
   2002 \tab - \tab 1 \tab 26 \tab 62 \tab 43 \tab 4 \tab 1 \tab 15 \tab 21 \tab 18 \tab 149 \tab 41 \tab 29 \tab 64 \tab 47 \tab 69 \tab 94 \tab 9 \tab 14 \tab 78 \tab 265 \tab 20 \tab 57 \tab 29 \tab 121 \tab 38 \tab 10 \tab 109 \tab 15 \tab 106 \tab 34 \tab 220 \tab 6 \tab 18 \tab 30 \tab 2 \tab 1 \tab 22 \tab 34 \tab 8 \tab 6 \tab 3 \tab 54 \tab 25 \tab 19 \tab 5 \tab 37 \tab 1 \tab 6 \tab 10 \tab 14 \tab 21 \tab 12 \tab 12 \tab 600 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 2765 \cr
   2004 \tab - \tab 3 \tab 36 \tab 62 \tab 62 \tab 3 \tab 5 \tab 9 \tab 22 \tab 19 \tab 129 \tab 37 \tab 42 \tab 51 \tab 56 \tab 91 \tab 95 \tab 12 \tab 18 \tab 81 \tab 246 \tab 15 \tab 60 \tab 29 \tab 105 \tab 52 \tab 6 \tab 114 \tab 12 \tab 92 \tab 25 \tab 202 \tab 5 \tab 22 \tab 26 \tab 1 \tab 5 \tab 22 \tab 24 \tab 17 \tab 7 \tab 7 \tab 65 \tab 31 \tab 26 \tab - \tab 30 \tab 1 \tab 9 \tab 8 \tab 19 \tab 34 \tab 18 \tab 20 \tab 611 \tab 3 \tab 3 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 2812 \cr
   2006 \tab - \tab 6 \tab 40 \tab 63 \tab 68 \tab 7 \tab 4 \tab 7 \tab 20 \tab 19 \tab 140 \tab 46 \tab 50 \tab 59 \tab 62 \tab 81 \tab 121 \tab 19 \tab 20 \tab 75 \tab 277 \tab 25 \tab 90 \tab 22 \tab 102 \tab 57 \tab 8 \tab 85 \tab 17 \tab 98 \tab 38 \tab 222 \tab 8 \tab 17 \tab 58 \tab 2 \tab 2 \tab 8 \tab 20 \tab 15 \tab 11 \tab 5 \tab 68 \tab 30 \tab 21 \tab 8 \tab 40 \tab 3 \tab 17 \tab 9 \tab 19 \tab 16 \tab 19 \tab 20 \tab 2141 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 6 \tab 34 \tab 27 \tab 48 \tab 5 \tab 1 \tab 13 \tab 15 \tab 15 \tab 72 \tab 26 \tab 27 \tab 67 \tab 43 \tab 55 \tab 70 \tab 8 \tab 13 \tab 47 \tab 189 \tab 15 \tab 44 \tab 29 \tab 58 \tab 47 \tab 3 \tab 78 \tab 7 \tab 76 \tab 25 \tab 142 \tab 7 \tab 23 \tab 30 \tab 2 \tab 1 \tab 15 \tab 21 \tab 10 \tab 5 \tab 8 \tab 52 \tab 23 \tab 13 \tab 4 \tab 20 \tab 2 \tab 1 \tab 8 \tab 10 \tab 13 \tab 9 \tab 9 \tab 422 \tab 5 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2023 \cr
   2010 \tab - \tab 3 \tab 30 \tab 33 \tab 27 \tab 4 \tab 4 \tab 11 \tab 18 \tab 15 \tab 104 \tab 22 \tab 53 \tab 41 \tab 43 \tab 51 \tab 64 \tab 12 \tab 11 \tab 49 \tab 187 \tab 16 \tab 47 \tab 14 \tab 59 \tab 44 \tab 5 \tab 70 \tab 11 \tab 62 \tab 29 \tab 175 \tab 6 \tab 11 \tab 28 \tab 1 \tab 3 \tab 17 \tab 22 \tab 8 \tab 4 \tab 4 \tab 42 \tab 20 \tab 18 \tab 5 \tab 24 \tab - \tab 8 \tab 11 \tab 10 \tab 23 \tab 9 \tab 21 \tab 432 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 21875 \tab 49 \tab 397 \tab 522 \tab 758 \tab 69 \tab 58 \tab 179 \tab 234 \tab 207 \tab 1475 \tab 409 \tab 476 \tab 542 \tab 685 \tab 893 \tab 1141 \tab 153 \tab 176 \tab 855 \tab 3620 \tab 235 \tab 839 \tab 351 \tab 1133 \tab 569 \tab 96 \tab 1177 \tab 202 \tab 929 \tab 401 \tab 2467 \tab 80 \tab 228 \tab 344 \tab 23 \tab 29 \tab 185 \tab 304 \tab 167 \tab 78 \tab 59 \tab 653 \tab 291 \tab 215 \tab 51 \tab 397 \tab 22 \tab 95 \tab 122 \tab 180 \tab 223 \tab 168 \tab 179 \tab 21479 \tab 32 \tab 20 \tab 16 \tab 10 \tab 11 \tab 13 \tab 3544 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
