% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_14.R
\name{parborn}
\alias{parborn}
\title{Were r's parents born in this country}
\source{
General Social Survey https://gss.norc.org
}
\description{
parborn
}
\details{
Question 32. Were both your parents born in this country?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab both born in the u.s. \tab mother don't know, father don't know \tab mother no, father yes \tab mother yes, father don't know \tab mother yes, father no \tab neither born in the u.s. \tab no answer \tab mother no, father don't know \tab mother don't know, father no \tab mother don't know, father yes \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 1196 \tab 4 \tab 35 \tab 1 \tab 65 \tab 228 \tab 1 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 1252 \tab 2 \tab 36 \tab 6 \tab 58 \tab 174 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 1178 \tab 3 \tab 30 \tab 3 \tab 58 \tab 193 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 1537 \tab 2 \tab 41 \tab 2 \tab 55 \tab 218 \tab 5 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 1262 \tab - \tab 44 \tab 1 \tab 59 \tab 229 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 1235 \tab 3 \tab 20 \tab - \tab 64 \tab 142 \tab 9 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 1265 \tab 2 \tab 34 \tab 3 \tab 58 \tab 168 \tab 4 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 1203 \tab - \tab 33 \tab 4 \tab 41 \tab 178 \tab 9 \tab - \tab 1 \tab 1 \tab - \tab 1470 \cr
   1987 \tab - \tab 1559 \tab 3 \tab 37 \tab 2 \tab 45 \tab 161 \tab 10 \tab 2 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 1235 \tab 2 \tab 26 \tab 3 \tab 46 \tab 163 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 1481 \cr
   1989 \tab - \tab 1266 \tab 1 \tab 40 \tab 3 \tab 50 \tab 168 \tab 8 \tab - \tab - \tab 1 \tab - \tab 1537 \cr
   1990 \tab - \tab 1134 \tab 1 \tab 28 \tab 4 \tab 52 \tab 147 \tab 6 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 1253 \tab 3 \tab 40 \tab 2 \tab 64 \tab 144 \tab 9 \tab 1 \tab 1 \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 1351 \tab 5 \tab 33 \tab 3 \tab 36 \tab 167 \tab 10 \tab 1 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 2480 \tab 3 \tab 71 \tab 8 \tab 91 \tab 318 \tab 17 \tab 2 \tab - \tab 2 \tab - \tab 2992 \cr
   1996 \tab - \tab 2441 \tab 3 \tab 52 \tab 6 \tab 88 \tab 303 \tab 11 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 2323 \tab 2 \tab 47 \tab 11 \tab 86 \tab 338 \tab 20 \tab 1 \tab 3 \tab 1 \tab - \tab 2832 \cr
   2000 \tab - \tab 2306 \tab 4 \tab 75 \tab 5 \tab 86 \tab 332 \tab 6 \tab 3 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 2273 \tab - \tab 72 \tab 4 \tab 86 \tab 314 \tab 12 \tab 3 \tab - \tab 1 \tab - \tab 2765 \cr
   2004 \tab - \tab 2295 \tab 1 \tab 61 \tab 2 \tab 80 \tab 359 \tab 10 \tab 3 \tab 1 \tab - \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 2360 \tab 1 \tab 55 \tab 6 \tab 91 \tab 467 \tab 5 \tab 5 \tab 1 \tab 1 \tab - \tab 4510 \cr
   2008 \tab - \tab 1598 \tab - \tab 34 \tab 3 \tab 59 \tab 320 \tab 6 \tab 2 \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 1631 \tab 1 \tab 44 \tab 2 \tab 40 \tab 314 \tab 10 \tab 1 \tab - \tab 1 \tab - \tab 2044 \cr
   2012 \tab - \tab 1504 \tab 1 \tab 59 \tab 8 \tab 54 \tab 336 \tab 9 \tab 2 \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 1950 \tab - \tab 66 \tab 6 \tab 69 \tab 435 \tab 4 \tab 2 \tab 3 \tab 3 \tab - \tab 2538 \cr
   2016 \tab - \tab 2238 \tab 2 \tab 84 \tab 5 \tab 92 \tab 429 \tab 13 \tab 2 \tab 1 \tab 1 \tab - \tab 2867 \cr
   2018 \tab - \tab 1823 \tab 4 \tab 62 \tab 7 \tab 72 \tab 371 \tab 6 \tab 3 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 12652 \tab 45148 \tab 53 \tab 1259 \tab 110 \tab 1745 \tab 7116 \tab 210 \tab 36 \tab 16 \tab 13 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} both born in the u.s.
\item \code{1} mother yes, father no
\item \code{2} mother no, father yes
\item \code{3} mother yes, father don't know
\item \code{4} mother no, father don't know
\item \code{5} mother don't know, father yes
\item \code{6} mother don't know, father no
\item \code{7} mother don't know, father don't know
\item \code{8} neither born in the u.s.
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
