% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_65.R
\name{parsol}
\alias{parsol}
\title{parsol}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's living standard compared to parents
}
\details{
Question 198a. Compared to your parents when they were the age you are now, do you think your own standard of living now is much better, somewhat better, about the same, somewhat worse, or much worse than theirs was?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab about the same \tab don't know \tab much better \tab much worse \tab no answer \tab somewhat better \tab somewhat worse \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1492 \tab 313 \tab 26 \tab 460 \tab 47 \tab 6 \tab 478 \tab 170 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 401 \tab 34 \tab 619 \tab 74 \tab 4 \tab 564 \tab 229 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 408 \tab 31 \tab 626 \tab 57 \tab 1 \tab 589 \tab 199 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 396 \tab 37 \tab 646 \tab 57 \tab 3 \tab 579 \tab 178 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 175 \tab 12 \tab 303 \tab 24 \tab 1 \tab 298 \tab 99 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 173 \tab 9 \tab 329 \tab 30 \tab 1 \tab 251 \tab 83 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 421 \tab 23 \tab 671 \tab 52 \tab 2 \tab 624 \tab 196 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 290 \tab 25 \tab 417 \tab 72 \tab 1 \tab 394 \tab 166 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 341 \tab 15 \tab 389 \tab 64 \tab 1 \tab 394 \tab 173 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 296 \tab 29 \tab 413 \tab 69 \tab 1 \tab 368 \tab 162 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 406 \tab 22 \tab 505 \tab 79 \tab 2 \tab 478 \tab 195 \tab - \tab 2538 \cr
   2016 \tab 912 \tab 466 \tab 20 \tab 573 \tab 100 \tab 3 \tab 555 \tab 238 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 362 \tab 24 \tab 495 \tab 65 \tab 3 \tab 433 \tab 181 \tab - \tab 2348 \cr
   2021 \tab 1360 \tab 703 \tab 4 \tab 656 \tab 106 \tab - \tab 835 \tab 353 \tab 15 \tab 4032 \cr
   2022 \tab 1173 \tab 551 \tab 31 \tab 628 \tab 167 \tab 6 \tab 697 \tab 286 \tab 5 \tab 3544 \cr
   Total \tab 47053 \tab 5702 \tab 342 \tab 7730 \tab 1063 \tab 35 \tab 7537 \tab 2908 \tab 20 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} much better
\item \code{2} somewhat better
\item \code{3} about the same
\item \code{4} somewhat worse
\item \code{5} much worse
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
