% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_551.R
\name{partners}
\alias{partners}
\title{How many sex partner's r had in last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
partners
}
\details{
Question 1541. How many sex partners have you had in the last 12 months?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 or more, (unspecified) \tab 1 partner \tab 11-20 partners \tab 2 partners \tab 21-100 partners \tab 3 partners \tab 4 partners \tab 5-10 partners \tab more than 100 partners \tab no partners \tab no answer \tab don't know \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 91 \tab 9 \tab 875 \tab 5 \tab 78 \tab 5 \tab 51 \tab 26 \tab 22 \tab 1 \tab 318 \tab - \tab - \tab 1481 \cr
   1989 \tab 136 \tab 6 \tab 906 \tab 2 \tab 89 \tab 2 \tab 40 \tab 20 \tab 23 \tab - \tab 308 \tab 5 \tab - \tab 1537 \cr
   1990 \tab 199 \tab 15 \tab 765 \tab 1 \tab 66 \tab 2 \tab 32 \tab 22 \tab 25 \tab 1 \tab 230 \tab 14 \tab - \tab 1372 \cr
   1991 \tab 872 \tab - \tab 418 \tab 2 \tab 39 \tab 1 \tab 15 \tab 8 \tab 7 \tab 1 \tab 146 \tab 8 \tab - \tab 1517 \cr
   1993 \tab 114 \tab - \tab 983 \tab 4 \tab 87 \tab 1 \tab 33 \tab 25 \tab 20 \tab - \tab 313 \tab 26 \tab - \tab 1606 \cr
   1994 \tab 201 \tab 3 \tab 1784 \tab 6 \tab 191 \tab 2 \tab 72 \tab 35 \tab 35 \tab - \tab 608 \tab 55 \tab - \tab 2992 \cr
   1996 \tab 258 \tab 25 \tab 1724 \tab 6 \tab 190 \tab 7 \tab 91 \tab 45 \tab 31 \tab - \tab 498 \tab 27 \tab 2 \tab 2904 \cr
   1998 \tab 383 \tab 13 \tab 1585 \tab 8 \tab 163 \tab 4 \tab 57 \tab 39 \tab 34 \tab 1 \tab 512 \tab 32 \tab 1 \tab 2832 \cr
   2000 \tab 417 \tab 2 \tab 1502 \tab 9 \tab 151 \tab - \tab 79 \tab 42 \tab 36 \tab - \tab 521 \tab 58 \tab - \tab 2817 \cr
   2002 \tab 489 \tab 5 \tab 1397 \tab 10 \tab 159 \tab 2 \tab 83 \tab 39 \tab 34 \tab 2 \tab 511 \tab 30 \tab 4 \tab 2765 \cr
   2004 \tab 584 \tab 8 \tab 1462 \tab 12 \tab 142 \tab 4 \tab 65 \tab 45 \tab 34 \tab 1 \tab 425 \tab 26 \tab 4 \tab 2812 \cr
   2006 \tab 2096 \tab 6 \tab 1485 \tab 17 \tab 152 \tab 3 \tab 66 \tab 37 \tab 38 \tab - \tab 596 \tab 13 \tab 1 \tab 4510 \cr
   2008 \tab 240 \tab 10 \tab 1092 \tab 11 \tab 120 \tab 3 \tab 55 \tab 25 \tab 33 \tab 2 \tab 415 \tab 13 \tab 4 \tab 2023 \cr
   2010 \tab 202 \tab 15 \tab 1129 \tab 2 \tab 124 \tab 1 \tab 63 \tab 39 \tab 25 \tab 2 \tab 427 \tab 14 \tab 1 \tab 2044 \cr
   2012 \tab 238 \tab 18 \tab 1068 \tab 5 \tab 109 \tab 4 \tab 57 \tab 34 \tab 23 \tab - \tab 405 \tab 13 \tab - \tab 1974 \cr
   2014 \tab 188 \tab 15 \tab 1490 \tab 7 \tab 130 \tab 1 \tab 68 \tab 32 \tab 31 \tab - \tab 537 \tab 37 \tab 2 \tab 2538 \cr
   2016 \tab 1096 \tab 9 \tab 1085 \tab 11 \tab 124 \tab 1 \tab 50 \tab 34 \tab 38 \tab - \tab 409 \tab 9 \tab 1 \tab 2867 \cr
   2018 \tab 942 \tab 4 \tab 868 \tab 1 \tab 79 \tab 1 \tab 40 \tab 21 \tab 23 \tab - \tab 358 \tab 10 \tab 1 \tab 2348 \cr
   2021 \tab 1699 \tab 6 \tab 1504 \tab 4 \tab 83 \tab 2 \tab 44 \tab 27 \tab 40 \tab 4 \tab 599 \tab 11 \tab 9 \tab 4032 \cr
   2022 \tab 1898 \tab - \tab 937 \tab 5 \tab 70 \tab 1 \tab 39 \tab 32 \tab 22 \tab 3 \tab 468 \tab 44 \tab 25 \tab 3544 \cr
   Total \tab 34218 \tab 169 \tab 24059 \tab 128 \tab 2346 \tab 47 \tab 1100 \tab 627 \tab 574 \tab 18 \tab 8604 \tab 445 \tab 55 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} no partners
\item \code{1} 1 partner
\item \code{2} 2 partners
\item \code{3} 3 partners
\item \code{4} 4 partners
\item \code{5} 5-10 partners
\item \code{6} 11-20 partners
\item \code{7} 21-100 partners
\item \code{8} more than 100 partners
\item \code{9} 1 or more, (unspecified)
\item \code{95} several
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
