% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_19.R
\name{partyid}
\alias{partyid}
\title{partyid}
\source{
General Social Survey https://gss.norc.org
}
\description{
Political party affiliation
}
\details{
Question 56. Generally speaking, do you usually think of yourself as a Republican, Democrat, Independent, or what?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab independent (neither, no response) \tab independent, close to democrat \tab independent, close to republican \tab no answer \tab not very strong democrat \tab not very strong republican \tab other party \tab strong democrat \tab strong republican \tab don't know \tab Total \cr
   1972 \tab 159 \tab 160 \tab 100 \tab 6 \tab 436 \tab 231 \tab 69 \tab 326 \tab 126 \tab - \tab 1613 \cr
   1973 \tab 143 \tab 192 \tab 141 \tab 11 \tab 383 \tab 218 \tab 62 \tab 231 \tab 123 \tab - \tab 1504 \cr
   1974 \tab 144 \tab 207 \tab 106 \tab 13 \tab 373 \tab 216 \tab 59 \tab 246 \tab 110 \tab 10 \tab 1484 \cr
   1975 \tab 211 \tab 211 \tab 122 \tab 5 \tab 349 \tab 236 \tab 12 \tab 251 \tab 93 \tab - \tab 1490 \cr
   1976 \tab 241 \tab 207 \tab 105 \tab 4 \tab 403 \tab 214 \tab 4 \tab 225 \tab 96 \tab - \tab 1499 \cr
   1977 \tab 175 \tab 200 \tab 130 \tab 12 \tab 400 \tab 226 \tab 5 \tab 275 \tab 107 \tab - \tab 1530 \cr
   1978 \tab 221 \tab 199 \tab 135 \tab 5 \tab 389 \tab 246 \tab 10 \tab 216 \tab 111 \tab - \tab 1532 \cr
   1980 \tab 246 \tab 193 \tab 123 \tab 3 \tab 374 \tab 218 \tab 9 \tab 188 \tab 114 \tab - \tab 1468 \cr
   1982 \tab 226 \tab 243 \tab 152 \tab 9 \tab 484 \tab 214 \tab 13 \tab 382 \tab 137 \tab - \tab 1860 \cr
   1983 \tab 192 \tab 218 \tab 141 \tab 6 \tab 382 \tab 256 \tab 13 \tab 246 \tab 145 \tab - \tab 1599 \cr
   1984 \tab 163 \tab 209 \tab 156 \tab 8 \tab 280 \tab 245 \tab 22 \tab 265 \tab 125 \tab - \tab 1473 \cr
   1985 \tab 146 \tab 159 \tab 157 \tab 5 \tab 351 \tab 266 \tab 17 \tab 244 \tab 189 \tab - \tab 1534 \cr
   1986 \tab 187 \tab 158 \tab 150 \tab 3 \tab 336 \tab 241 \tab 9 \tab 245 \tab 141 \tab - \tab 1470 \cr
   1987 \tab 199 \tab 211 \tab 148 \tab 10 \tab 395 \tab 266 \tab 15 \tab 422 \tab 153 \tab - \tab 1819 \cr
   1988 \tab 186 \tab 180 \tab 140 \tab - \tab 313 \tab 275 \tab 5 \tab 236 \tab 146 \tab - \tab 1481 \cr
   1989 \tab 192 \tab 134 \tab 119 \tab 5 \tab 335 \tab 330 \tab 13 \tab 234 \tab 175 \tab - \tab 1537 \cr
   1990 \tab 155 \tab 133 \tab 143 \tab 4 \tab 315 \tab 280 \tab 12 \tab 170 \tab 160 \tab - \tab 1372 \cr
   1991 \tab 188 \tab 131 \tab 169 \tab 6 \tab 319 \tab 285 \tab 14 \tab 224 \tab 181 \tab - \tab 1517 \cr
   1993 \tab 205 \tab 190 \tab 158 \tab 9 \tab 321 \tab 299 \tab 17 \tab 227 \tab 180 \tab - \tab 1606 \cr
   1994 \tab 369 \tab 341 \tab 282 \tab 49 \tab 644 \tab 519 \tab 44 \tab 423 \tab 321 \tab - \tab 2992 \cr
   1996 \tab 457 \tab 356 \tab 258 \tab 6 \tab 577 \tab 500 \tab 43 \tab 400 \tab 307 \tab - \tab 2904 \cr
   1998 \tab 477 \tab 349 \tab 244 \tab 9 \tab 597 \tab 484 \tab 63 \tab 370 \tab 239 \tab - \tab 2832 \cr
   2000 \tab 566 \tab 325 \tab 261 \tab 12 \tab 507 \tab 399 \tab 48 \tab 414 \tab 285 \tab - \tab 2817 \cr
   2002 \tab 528 \tab 267 \tab 199 \tab 36 \tab 515 \tab 449 \tab 48 \tab 408 \tab 315 \tab - \tab 2765 \cr
   2004 \tab 471 \tab 281 \tab 239 \tab 12 \tab 504 \tab 425 \tab 29 \tab 455 \tab 396 \tab - \tab 2812 \cr
   2006 \tab 997 \tab 527 \tab 327 \tab 26 \tab 736 \tab 637 \tab 65 \tab 700 \tab 495 \tab - \tab 4510 \cr
   2008 \tab 322 \tab 262 \tab 162 \tab 13 \tab 331 \tab 303 \tab 38 \tab 390 \tab 202 \tab - \tab 2023 \cr
   2010 \tab 360 \tab 265 \tab 197 \tab 16 \tab 348 \tab 277 \tab 49 \tab 348 \tab 184 \tab - \tab 2044 \cr
   2012 \tab 373 \tab 235 \tab 157 \tab 14 \tab 343 \tab 250 \tab 54 \tab 356 \tab 192 \tab - \tab 1974 \cr
   2014 \tab 502 \tab 337 \tab 249 \tab 25 \tab 406 \tab 292 \tab 62 \tab 419 \tab 245 \tab 1 \tab 2538 \cr
   2016 \tab 473 \tab 405 \tab 292 \tab 32 \tab 496 \tab 364 \tab 72 \tab 463 \tab 270 \tab - \tab 2867 \cr
   2018 \tab 414 \tab 307 \tab 259 \tab 33 \tab 352 \tab 272 \tab 77 \tab 379 \tab 255 \tab - \tab 2348 \cr
   2021 \tab 817 \tab 471 \tab 327 \tab 32 \tab 541 \tab 384 \tab 114 \tab 822 \tab 524 \tab - \tab 4032 \cr
   2022 \tab 835 \tab 400 \tab 330 \tab 35 \tab 451 \tab 361 \tab 106 \tab 595 \tab 431 \tab - \tab 3544 \cr
   Total \tab 11540 \tab 8663 \tab 6378 \tab 474 \tab 14286 \tab 10678 \tab 1292 \tab 11795 \tab 7273 \tab 11 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} strong democrat
\item \code{1} not very strong democrat
\item \code{2} independent, close to democrat
\item \code{3} independent (neither, no response)
\item \code{4} independent, close to republican
\item \code{5} not very strong republican
\item \code{6} strong republican
\item \code{7} other party
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
