% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_597.R
\name{pasei}
\alias{pasei}
\title{R's father's socioeconomic index (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
pasei
}
\details{
Question 1648. Socioeconomic index of respondent's father.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 17.1 \tab 17.6 \tab 19.4 \tab 20.1 \tab 21.1 \tab 22.7 \tab 22.9 \tab 24.2 \tab 24.6 \tab 24.8 \tab 25.5 \tab 25.9 \tab 26.4 \tab 26.5 \tab 26.6 \tab 26.7 \tab 26.9 \tab 27.5 \tab 27.7 \tab 28 \tab 28.2 \tab 28.4 \tab 28.6 \tab 29.2 \tab 29.3 \tab 29.4 \tab 29.5 \tab 29.6 \tab 29.7 \tab 29.8 \tab 30.1 \tab 30.2 \tab 30.4 \tab 30.5 \tab 30.6 \tab 31.3 \tab 31.4 \tab 31.5 \tab 31.7 \tab 31.9 \tab 32 \tab 32.1 \tab 32.3 \tab 32.7 \tab 32.8 \tab 33.1 \tab 33.2 \tab 33.3 \tab 33.4 \tab 33.5 \tab 33.6 \tab 33.9 \tab 34.2 \tab 34.3 \tab 35.1 \tab 35.4 \tab 35.7 \tab 36.4 \tab 36.5 \tab 36.7 \tab 36.9 \tab 37.2 \tab 37.3 \tab 37.4 \tab 37.5 \tab 37.6 \tab 37.7 \tab 37.8 \tab 37.9 \tab 38 \tab 38.1 \tab 38.2 \tab 38.4 \tab 38.5 \tab 38.8 \tab 38.9 \tab 39 \tab 39.4 \tab 39.5 \tab 40 \tab 41.3 \tab 41.6 \tab 42.2 \tab 42.4 \tab 42.7 \tab 43.1 \tab 43.4 \tab 44.2 \tab 44.7 \tab 44.8 \tab 45.2 \tab 45.3 \tab 45.4 \tab 45.8 \tab 45.9 \tab 46.3 \tab 48.5 \tab 48.6 \tab 48.8 \tab 48.9 \tab 49.3 \tab 49.4 \tab 49.6 \tab 49.9 \tab 50.2 \tab 50.3 \tab 50.5 \tab 50.7 \tab 51 \tab 51.9 \tab 52.3 \tab 52.4 \tab 52.5 \tab 53.1 \tab 53.4 \tab 53.7 \tab 53.9 \tab 54.2 \tab 54.5 \tab 54.7 \tab 54.9 \tab 55.1 \tab 55.3 \tab 57.1 \tab 58.3 \tab 59.3 \tab 60.5 \tab 61.9 \tab 62 \tab 62.1 \tab 62.5 \tab 63.2 \tab 63.5 \tab 64.1 \tab 64.2 \tab 64.4 \tab 64.8 \tab 65.8 \tab 66.1 \tab 66.3 \tab 67 \tab 68.1 \tab 69.2 \tab 69.7 \tab 71.3 \tab 73.3 \tab 73.6 \tab 73.9 \tab 74 \tab 74.4 \tab 74.9 \tab 75.2 \tab 76.4 \tab 78.5 \tab 79.9 \tab 80.3 \tab 80.9 \tab 82.7 \tab 83.2 \tab 83.7 \tab 83.9 \tab 84.7 \tab 85.1 \tab 86 \tab 86.5 \tab 86.6 \tab 86.8 \tab 87.2 \tab 87.3 \tab 87.4 \tab 87.9 \tab 88.3 \tab 88.7 \tab 88.8 \tab 89.2 \tab 90.1 \tab 90.8 \tab 92.3 \tab 92.8 \tab 93.3 \tab 96 \tab 97.2 \tab 18.5 \tab 19.5 \tab 25.2 \tab 25.3 \tab 25.4 \tab 25.8 \tab 27.2 \tab 27.3 \tab 27.4 \tab 27.9 \tab 28.8 \tab 29.1 \tab 30 \tab 32.4 \tab 32.9 \tab 33 \tab 36 \tab 36.2 \tab 37 \tab 37.1 \tab 39.7 \tab 39.9 \tab 40.9 \tab 42.3 \tab 42.5 \tab 43.2 \tab 43.3 \tab 43.6 \tab 47 \tab 47.8 \tab 47.9 \tab 48.4 \tab 53.3 \tab 56.5 \tab 59 \tab 59.1 \tab 59.9 \tab 60.3 \tab 60.6 \tab 62.2 \tab 65.2 \tab 68.8 \tab 72 \tab 73.7 \tab 73.8 \tab 78 \tab 80 \tab 81 \tab 81.9 \tab 84.2 \tab 84.5 \tab 85.7 \tab 86.9 \tab 87 \tab 87.5 \tab 88.2 \tab 88.6 \tab 90.3 \tab 28.9 \tab 30.7 \tab 34 \tab 34.4 \tab 36.1 \tab 38.7 \tab 40.5 \tab 41.1 \tab 43.7 \tab 45.7 \tab 46.8 \tab 47.6 \tab 48.7 \tab 56.4 \tab 70 \tab 70.5 \tab 71.2 \tab 72.5 \tab 79.8 \tab 86.7 \tab 91.3 \tab 93.1 \tab 21.2 \tab 29 \tab 31.1 \tab 41.2 \tab 46.7 \tab 49.7 \tab 53.2 \tab 58.8 \tab 64 \tab 71.5 \tab 75.3 \tab 76.3 \tab 22.5 \tab 29.9 \tab 34.8 \tab 42.9 \tab 44.3 \tab 44.4 \tab 78.8 \tab 85 \tab 86.2 \tab 88.9 \tab 91.7 \tab 30.8 \tab 34.6 \tab 45 \tab 61.4 \tab 78.7 \tab 78.9 \tab 85.5 \tab 85.8 \tab 92.9 \tab 26.3 \tab 35.6 \tab 77.8 \tab 80.1 \tab 34.7 \tab 35.2 \tab 38.3 \tab 42 \tab 46 \tab 73.2 \tab 75.9 \tab 81.3 \tab 91.2 \tab 22.6 \tab 40.1 \tab 84 \tab 31.2 \tab 58.1 \tab 93.7 \tab 24.5 \tab 60.4 \tab 71.6 \tab 74.6 \tab 43.5 \tab 61.1 \tab 56.6 \tab 63.9 \tab 46.1 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 241 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 3 \tab 2 \tab 2 \tab 5 \tab 1 \tab 4 \tab 4 \tab 2 \tab 31 \tab 4 \tab 13 \tab 6 \tab 4 \tab 12 \tab 1 \tab 1 \tab 18 \tab 21 \tab 35 \tab 5 \tab 7 \tab 5 \tab 2 \tab 1 \tab 11 \tab 3 \tab 2 \tab 4 \tab 2 \tab 2 \tab 10 \tab 9 \tab 31 \tab 1 \tab 2 \tab 3 \tab 4 \tab 23 \tab 3 \tab 1 \tab 6 \tab 9 \tab 1 \tab 1 \tab 2 \tab 5 \tab 7 \tab 31 \tab 4 \tab 12 \tab 7 \tab 2 \tab 10 \tab 18 \tab 2 \tab 2 \tab 3 \tab 197 \tab 5 \tab 1 \tab 1 \tab 2 \tab 10 \tab 4 \tab 6 \tab 7 \tab 22 \tab 6 \tab 6 \tab 1 \tab 14 \tab 4 \tab 8 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 12 \tab 1 \tab 2 \tab 3 \tab 9 \tab 7 \tab 7 \tab 1 \tab 15 \tab 7 \tab 3 \tab 7 \tab 2 \tab 1 \tab 19 \tab 3 \tab 3 \tab 1 \tab 4 \tab 35 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab 4 \tab 5 \tab 7 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 3 \tab 5 \tab 5 \tab 5 \tab 12 \tab 116 \tab 11 \tab 2 \tab 1 \tab 1 \tab 7 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 8 \tab 2 \tab 5 \tab 2 \tab 10 \tab 1 \tab 1 \tab 3 \tab 3 \tab 10 \tab 2 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab 9 \tab 7 \tab 5 \tab 8 \tab 3 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 5 \tab 3 \tab 3 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 263 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 4 \tab 3 \tab 27 \tab 3 \tab 8 \tab 5 \tab 3 \tab 22 \tab - \tab 1 \tab 9 \tab 27 \tab 40 \tab 2 \tab 14 \tab 6 \tab 3 \tab - \tab 9 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 8 \tab 8 \tab 37 \tab - \tab 1 \tab - \tab - \tab 30 \tab 4 \tab - \tab 7 \tab 3 \tab 1 \tab 2 \tab - \tab 6 \tab 4 \tab 16 \tab 5 \tab 7 \tab 8 \tab - \tab 21 \tab 20 \tab - \tab 1 \tab 2 \tab 215 \tab 1 \tab 2 \tab 2 \tab 2 \tab 8 \tab - \tab 3 \tab 17 \tab 19 \tab 9 \tab 4 \tab - \tab 4 \tab 4 \tab 4 \tab 2 \tab 2 \tab - \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 8 \tab 1 \tab 5 \tab 5 \tab 2 \tab 7 \tab 7 \tab - \tab 28 \tab 5 \tab 1 \tab 3 \tab 3 \tab - \tab 11 \tab 3 \tab 3 \tab 2 \tab 7 \tab 29 \tab 6 \tab - \tab 1 \tab 6 \tab 3 \tab - \tab 2 \tab 6 \tab 10 \tab 7 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab 5 \tab 3 \tab 2 \tab 6 \tab 14 \tab 104 \tab 7 \tab 6 \tab - \tab - \tab 10 \tab 4 \tab - \tab - \tab 2 \tab 5 \tab 5 \tab - \tab 4 \tab 6 \tab 12 \tab - \tab 1 \tab 1 \tab 2 \tab 8 \tab 4 \tab - \tab 8 \tab 3 \tab 1 \tab 3 \tab 3 \tab 2 \tab - \tab 10 \tab 8 \tab 9 \tab 5 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 6 \tab 1 \tab 3 \tab - \tab 6 \tab 1 \tab 1 \tab 4 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 5 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 248 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 27 \tab 2 \tab 16 \tab 3 \tab 9 \tab 7 \tab 1 \tab 2 \tab 5 \tab 29 \tab 37 \tab 5 \tab 15 \tab 2 \tab 2 \tab - \tab 5 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 14 \tab 7 \tab 35 \tab 1 \tab 1 \tab 9 \tab - \tab 39 \tab 2 \tab 1 \tab 9 \tab 1 \tab 1 \tab - \tab 1 \tab 5 \tab 8 \tab 29 \tab 6 \tab 5 \tab 1 \tab 2 \tab 6 \tab 26 \tab - \tab - \tab - \tab 154 \tab 1 \tab - \tab - \tab 2 \tab 9 \tab 1 \tab 7 \tab 12 \tab 17 \tab 9 \tab 4 \tab 2 \tab 6 \tab 2 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 5 \tab - \tab - \tab - \tab 2 \tab 1 \tab 9 \tab - \tab 4 \tab 3 \tab 3 \tab 8 \tab 12 \tab 1 \tab 24 \tab 3 \tab 5 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 7 \tab 3 \tab 6 \tab 39 \tab 3 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab 9 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 4 \tab 4 \tab 5 \tab 14 \tab 87 \tab 17 \tab 5 \tab - \tab - \tab 9 \tab 4 \tab - \tab - \tab 2 \tab - \tab 8 \tab - \tab 5 \tab 1 \tab 8 \tab 2 \tab 2 \tab - \tab 3 \tab 8 \tab 3 \tab - \tab 6 \tab 1 \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 4 \tab 10 \tab 5 \tab 8 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 5 \tab 2 \tab 5 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 292 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 5 \tab - \tab - \tab 2 \tab 4 \tab 3 \tab 16 \tab 3 \tab 9 \tab 1 \tab 4 \tab 10 \tab 2 \tab 4 \tab 19 \tab 18 \tab 30 \tab 3 \tab 16 \tab 10 \tab 2 \tab - \tab 10 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 10 \tab 9 \tab 39 \tab 1 \tab 1 \tab 5 \tab 1 \tab 30 \tab 4 \tab 1 \tab 6 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 6 \tab 32 \tab 6 \tab 10 \tab 6 \tab 1 \tab 20 \tab 18 \tab 2 \tab - \tab 3 \tab 170 \tab 4 \tab 2 \tab - \tab 5 \tab 5 \tab 3 \tab 5 \tab 15 \tab 11 \tab 7 \tab 1 \tab 2 \tab 4 \tab 2 \tab 2 \tab 2 \tab 2 \tab 3 \tab - \tab 8 \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab 14 \tab 1 \tab 32 \tab - \tab 2 \tab 4 \tab 10 \tab - \tab 20 \tab 6 \tab 1 \tab 7 \tab 2 \tab - \tab 16 \tab 1 \tab 3 \tab 1 \tab 4 \tab 45 \tab 1 \tab 2 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab 13 \tab 3 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 4 \tab 2 \tab 7 \tab 6 \tab 84 \tab 26 \tab 6 \tab - \tab 1 \tab 10 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 6 \tab - \tab 6 \tab 3 \tab 8 \tab 2 \tab 2 \tab 2 \tab 1 \tab 12 \tab 2 \tab - \tab 8 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 8 \tab 7 \tab 5 \tab 13 \tab 4 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 4 \tab - \tab 5 \tab 4 \tab 10 \tab - \tab - \tab 1 \tab 6 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 255 \tab 2 \tab - \tab - \tab 6 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 16 \tab 1 \tab 11 \tab 5 \tab 3 \tab 1 \tab 2 \tab 2 \tab 17 \tab 15 \tab 46 \tab 3 \tab 20 \tab 3 \tab 2 \tab - \tab 7 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 11 \tab 13 \tab 38 \tab 2 \tab 2 \tab 6 \tab 2 \tab 40 \tab 7 \tab 2 \tab 8 \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab 31 \tab 8 \tab 6 \tab 7 \tab - \tab 2 \tab 29 \tab 1 \tab - \tab 1 \tab 218 \tab 4 \tab - \tab 1 \tab 2 \tab 8 \tab - \tab 4 \tab 21 \tab 20 \tab 11 \tab 7 \tab 1 \tab 4 \tab 5 \tab 3 \tab - \tab - \tab - \tab - \tab 6 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 15 \tab - \tab 5 \tab 4 \tab 6 \tab 2 \tab 16 \tab 2 \tab 28 \tab 6 \tab 3 \tab 6 \tab - \tab 1 \tab 6 \tab 4 \tab 10 \tab 1 \tab 2 \tab 66 \tab 10 \tab 1 \tab 2 \tab 10 \tab 2 \tab 1 \tab 1 \tab - \tab 8 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab 6 \tab 20 \tab 94 \tab 29 \tab 4 \tab - \tab - \tab 13 \tab 5 \tab 1 \tab 3 \tab - \tab 2 \tab 6 \tab 1 \tab 2 \tab 2 \tab 11 \tab 1 \tab 2 \tab 1 \tab 2 \tab 12 \tab 6 \tab 1 \tab 9 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 6 \tab 4 \tab 11 \tab 7 \tab 6 \tab - \tab 1 \tab 1 \tab 9 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 10 \tab - \tab 3 \tab 4 \tab 13 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 4 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 5 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 592 \tab 1 \tab - \tab 5 \tab 4 \tab - \tab 1 \tab 3 \tab 12 \tab 2 \tab - \tab 5 \tab 4 \tab 3 \tab 18 \tab 5 \tab 28 \tab 9 \tab 11 \tab 11 \tab 2 \tab 5 \tab 34 \tab 46 \tab 87 \tab 3 \tab 25 \tab 7 \tab 6 \tab 2 \tab 15 \tab 7 \tab - \tab 4 \tab - \tab - \tab 21 \tab 16 \tab 86 \tab - \tab 4 \tab 8 \tab 4 \tab 61 \tab 9 \tab 5 \tab 13 \tab 9 \tab 2 \tab 3 \tab 4 \tab 8 \tab 8 \tab 47 \tab 13 \tab 5 \tab 20 \tab 5 \tab 19 \tab 75 \tab 1 \tab - \tab 2 \tab 335 \tab 10 \tab 3 \tab 6 \tab 6 \tab 14 \tab - \tab 9 \tab 20 \tab 31 \tab 22 \tab 5 \tab 1 \tab 3 \tab 9 \tab 14 \tab 2 \tab - \tab 2 \tab 2 \tab 18 \tab - \tab - \tab 1 \tab 3 \tab 4 \tab 21 \tab 1 \tab 1 \tab 2 \tab 11 \tab 10 \tab 8 \tab 2 \tab 44 \tab 11 \tab 5 \tab 9 \tab 1 \tab 2 \tab 18 \tab 5 \tab 10 \tab 6 \tab 17 \tab 94 \tab 8 \tab 7 \tab 3 \tab 13 \tab 3 \tab 3 \tab 3 \tab 3 \tab 14 \tab 5 \tab - \tab 2 \tab 2 \tab 5 \tab 2 \tab 2 \tab 5 \tab 1 \tab 2 \tab 9 \tab 7 \tab 6 \tab 6 \tab 23 \tab 161 \tab 43 \tab 7 \tab - \tab - \tab 19 \tab 3 \tab 1 \tab 4 \tab - \tab 1 \tab 8 \tab - \tab 14 \tab 2 \tab 20 \tab 3 \tab 4 \tab 2 \tab 2 \tab 26 \tab 12 \tab 1 \tab 12 \tab 2 \tab 1 \tab 1 \tab 7 \tab 6 \tab 1 \tab 16 \tab 9 \tab 10 \tab 18 \tab 7 \tab 2 \tab - \tab - \tab 15 \tab - \tab - \tab 6 \tab 2 \tab - \tab 2 \tab 24 \tab 11 \tab 7 \tab 6 \tab 12 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab 2 \tab - \tab - \tab - \tab 3 \tab 2 \tab 2 \tab - \tab - \tab - \tab 2 \tab 6 \tab 6 \tab 1 \tab 4 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 5 \tab 1 \tab 3 \tab 2 \tab - \tab 4 \tab 1 \tab 8 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 578 \tab 3 \tab 3 \tab 2 \tab 4 \tab 1 \tab 3 \tab 6 \tab 7 \tab 2 \tab 1 \tab 1 \tab 5 \tab 5 \tab 34 \tab 5 \tab 22 \tab 6 \tab 7 \tab 23 \tab - \tab 3 \tab 32 \tab 64 \tab 88 \tab 10 \tab 24 \tab 11 \tab 2 \tab - \tab 14 \tab 5 \tab 4 \tab 3 \tab 1 \tab 2 \tab 14 \tab 13 \tab 76 \tab 2 \tab 2 \tab 5 \tab 3 \tab 62 \tab 13 \tab 6 \tab 12 \tab 9 \tab 2 \tab - \tab 3 \tab 3 \tab 8 \tab 53 \tab 3 \tab 19 \tab 13 \tab 7 \tab 13 \tab 33 \tab 1 \tab - \tab 1 \tab 256 \tab 5 \tab 1 \tab - \tab 5 \tab 11 \tab - \tab 10 \tab 19 \tab 40 \tab 12 \tab 5 \tab 5 \tab 22 \tab 5 \tab 9 \tab 2 \tab - \tab 2 \tab 1 \tab 6 \tab 2 \tab - \tab - \tab 6 \tab 3 \tab 34 \tab 3 \tab 11 \tab 5 \tab 8 \tab 13 \tab 9 \tab 2 \tab 47 \tab 12 \tab 3 \tab 10 \tab 3 \tab - \tab 25 \tab 3 \tab 11 \tab 3 \tab 12 \tab 72 \tab 5 \tab 1 \tab 5 \tab 8 \tab 6 \tab 3 \tab 4 \tab 7 \tab 14 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 9 \tab 6 \tab 18 \tab 24 \tab 187 \tab 31 \tab 9 \tab - \tab - \tab 17 \tab 9 \tab 1 \tab 4 \tab 2 \tab 6 \tab 27 \tab - \tab 11 \tab 10 \tab 23 \tab 5 \tab - \tab 3 \tab 1 \tab 22 \tab 15 \tab - \tab 20 \tab 1 \tab 2 \tab 1 \tab 4 \tab 4 \tab 2 \tab 12 \tab 13 \tab 15 \tab 27 \tab 5 \tab - \tab 1 \tab - \tab 5 \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 16 \tab 8 \tab 8 \tab 4 \tab 15 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 5 \tab 3 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 6 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 14 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 14 \tab 3 \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab 6 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 604 \tab 3 \tab 1 \tab 5 \tab 3 \tab - \tab 4 \tab 3 \tab 6 \tab 4 \tab 1 \tab 5 \tab 1 \tab 3 \tab 27 \tab 6 \tab 34 \tab 6 \tab 8 \tab 34 \tab 2 \tab 2 \tab 34 \tab 63 \tab 71 \tab 4 \tab 26 \tab 4 \tab 3 \tab 1 \tab 12 \tab 7 \tab 2 \tab 1 \tab - \tab 3 \tab 14 \tab 11 \tab 63 \tab - \tab - \tab 10 \tab 2 \tab 54 \tab 10 \tab 4 \tab 15 \tab 5 \tab 3 \tab - \tab 3 \tab 6 \tab 2 \tab 64 \tab 4 \tab 11 \tab 4 \tab 1 \tab 26 \tab 32 \tab - \tab - \tab 3 \tab 256 \tab - \tab 1 \tab 4 \tab 6 \tab 13 \tab 2 \tab 6 \tab 27 \tab 45 \tab 10 \tab 6 \tab 4 \tab 17 \tab 6 \tab 15 \tab 6 \tab 4 \tab - \tab 1 \tab 8 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 33 \tab 1 \tab 6 \tab 8 \tab 4 \tab 12 \tab 9 \tab 3 \tab 44 \tab 10 \tab 5 \tab 12 \tab 1 \tab 1 \tab 23 \tab 9 \tab 6 \tab 2 \tab 7 \tab 85 \tab 2 \tab 2 \tab 1 \tab 7 \tab 8 \tab 1 \tab - \tab 4 \tab 18 \tab 6 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 7 \tab 5 \tab 4 \tab 5 \tab 7 \tab 29 \tab 174 \tab 26 \tab 9 \tab - \tab 1 \tab 23 \tab 1 \tab - \tab 5 \tab 1 \tab 3 \tab 17 \tab - \tab 11 \tab 7 \tab 19 \tab 3 \tab 3 \tab - \tab 2 \tab 28 \tab 12 \tab - \tab 19 \tab 1 \tab 1 \tab 7 \tab 3 \tab 4 \tab 2 \tab 9 \tab 15 \tab 18 \tab 21 \tab 6 \tab - \tab 2 \tab 1 \tab 6 \tab - \tab - \tab 6 \tab 2 \tab 1 \tab - \tab 19 \tab 2 \tab 2 \tab 4 \tab 16 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab 3 \tab - \tab 4 \tab - \tab - \tab 4 \tab 2 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab 3 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 8 \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab 2 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 9 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 622 \tab 10 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 7 \tab 2 \tab 1 \tab 4 \tab 3 \tab 4 \tab 34 \tab 2 \tab 24 \tab 8 \tab 12 \tab 37 \tab 1 \tab 2 \tab 35 \tab 51 \tab 65 \tab 7 \tab 36 \tab 10 \tab 2 \tab - \tab 9 \tab 1 \tab 5 \tab 2 \tab 1 \tab 5 \tab 13 \tab 11 \tab 118 \tab 4 \tab 1 \tab 11 \tab 5 \tab 62 \tab 16 \tab 6 \tab 10 \tab 6 \tab 3 \tab - \tab 1 \tab 10 \tab 8 \tab 54 \tab 7 \tab - \tab 13 \tab 1 \tab 24 \tab 38 \tab 1 \tab - \tab 6 \tab 291 \tab 5 \tab 1 \tab 3 \tab 9 \tab 10 \tab 1 \tab 9 \tab 24 \tab 32 \tab 8 \tab 4 \tab 2 \tab 10 \tab 7 \tab 14 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab 26 \tab 6 \tab - \tab 3 \tab 2 \tab 7 \tab 7 \tab 5 \tab 54 \tab 12 \tab 3 \tab 2 \tab 5 \tab - \tab 13 \tab 4 \tab 12 \tab 2 \tab 5 \tab 67 \tab 1 \tab 1 \tab 3 \tab 16 \tab 2 \tab 2 \tab 1 \tab 8 \tab 7 \tab 6 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab 2 \tab 7 \tab 4 \tab 5 \tab 4 \tab 6 \tab 21 \tab 165 \tab 39 \tab 12 \tab - \tab 1 \tab 17 \tab 3 \tab - \tab 5 \tab - \tab 2 \tab 8 \tab - \tab 8 \tab 7 \tab 20 \tab 3 \tab 1 \tab 1 \tab 2 \tab 21 \tab 11 \tab - \tab 16 \tab 1 \tab 6 \tab 1 \tab 3 \tab 2 \tab 1 \tab 12 \tab 14 \tab 12 \tab 25 \tab 3 \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 3 \tab 15 \tab 3 \tab 9 \tab 4 \tab 13 \tab 2 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 5 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 5 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 2 \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 10 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 653 \tab 2 \tab 1 \tab 1 \tab 7 \tab - \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 3 \tab - \tab 2 \tab 56 \tab 4 \tab 20 \tab 8 \tab 13 \tab 52 \tab 2 \tab 2 \tab 26 \tab 46 \tab 40 \tab 6 \tab 32 \tab 8 \tab - \tab - \tab 10 \tab 9 \tab 2 \tab 2 \tab - \tab 2 \tab 11 \tab 13 \tab 77 \tab - \tab - \tab 11 \tab - \tab 58 \tab 9 \tab 2 \tab 11 \tab 9 \tab 7 \tab - \tab 1 \tab 2 \tab 8 \tab 54 \tab 9 \tab 16 \tab 14 \tab 1 \tab 17 \tab 50 \tab 2 \tab - \tab 5 \tab 207 \tab 3 \tab - \tab 5 \tab 6 \tab 8 \tab 1 \tab 10 \tab 23 \tab 22 \tab 9 \tab 4 \tab 2 \tab 8 \tab 3 \tab 11 \tab 3 \tab 1 \tab - \tab - \tab 5 \tab 6 \tab - \tab - \tab 2 \tab 3 \tab 30 \tab 1 \tab 1 \tab 3 \tab 7 \tab 9 \tab 17 \tab - \tab 63 \tab 1 \tab 4 \tab 8 \tab 1 \tab - \tab 21 \tab 10 \tab 4 \tab - \tab 7 \tab 96 \tab 1 \tab 2 \tab 4 \tab 9 \tab 5 \tab 2 \tab 5 \tab 1 \tab 17 \tab 7 \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab 3 \tab 3 \tab 12 \tab 32 \tab 171 \tab 40 \tab 8 \tab - \tab 1 \tab 13 \tab 8 \tab 1 \tab 5 \tab - \tab 3 \tab 11 \tab - \tab 14 \tab 9 \tab 16 \tab 1 \tab 2 \tab - \tab 2 \tab 25 \tab 10 \tab - \tab 17 \tab - \tab 2 \tab 1 \tab 7 \tab 2 \tab 2 \tab 20 \tab 12 \tab 18 \tab 12 \tab 7 \tab - \tab - \tab 3 \tab 4 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 12 \tab 6 \tab 6 \tab 4 \tab 12 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 6 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 7 \tab 1 \tab 2 \tab 3 \tab - \tab 2 \tab 4 \tab 3 \tab 5 \tab - \tab 2 \tab - \tab - \tab 1 \tab 5 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 653 \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 5 \tab 4 \tab 1 \tab 2 \tab 3 \tab 2 \tab 58 \tab 1 \tab 33 \tab 16 \tab 5 \tab 39 \tab 2 \tab 2 \tab 31 \tab 45 \tab 61 \tab 6 \tab 25 \tab 14 \tab - \tab - \tab 14 \tab 7 \tab 2 \tab 3 \tab - \tab - \tab 17 \tab 10 \tab 73 \tab 1 \tab 3 \tab 7 \tab 3 \tab 51 \tab 10 \tab 8 \tab 20 \tab 11 \tab 4 \tab 1 \tab 2 \tab 7 \tab 4 \tab 62 \tab 3 \tab 13 \tab 9 \tab 2 \tab 23 \tab 37 \tab 3 \tab - \tab 3 \tab 182 \tab 5 \tab - \tab 2 \tab 6 \tab 13 \tab 2 \tab 6 \tab 27 \tab 31 \tab 9 \tab 6 \tab 2 \tab 13 \tab 7 \tab 13 \tab 2 \tab 2 \tab 3 \tab 3 \tab 5 \tab 6 \tab - \tab 1 \tab 2 \tab - \tab 24 \tab 2 \tab 1 \tab 4 \tab 6 \tab 6 \tab 15 \tab - \tab 63 \tab 14 \tab 6 \tab 1 \tab 1 \tab - \tab 23 \tab 12 \tab 5 \tab 1 \tab 4 \tab 81 \tab - \tab - \tab 2 \tab 11 \tab 9 \tab 3 \tab 1 \tab 5 \tab 15 \tab 7 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 6 \tab 2 \tab 4 \tab 1 \tab 12 \tab 22 \tab 157 \tab 37 \tab 8 \tab 1 \tab 1 \tab 18 \tab 5 \tab - \tab 4 \tab 1 \tab 4 \tab 15 \tab 1 \tab 15 \tab 7 \tab 14 \tab 4 \tab 1 \tab 2 \tab - \tab 26 \tab 20 \tab - \tab 21 \tab 1 \tab 4 \tab 1 \tab 12 \tab 4 \tab 2 \tab 11 \tab 17 \tab 13 \tab 17 \tab 8 \tab - \tab - \tab 2 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 18 \tab 6 \tab 4 \tab 4 \tab 20 \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 3 \tab - \tab 4 \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 4 \tab 7 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 7 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 6 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 7 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2199 \tab 3 \tab 2 \tab 1 \tab 5 \tab - \tab 5 \tab 3 \tab 5 \tab 5 \tab 2 \tab 3 \tab 3 \tab 4 \tab 56 \tab 3 \tab 29 \tab 21 \tab 11 \tab 43 \tab - \tab 4 \tab 32 \tab 68 \tab 68 \tab 3 \tab 34 \tab 8 \tab 3 \tab - \tab 13 \tab 7 \tab 2 \tab 3 \tab - \tab 3 \tab 20 \tab 15 \tab 85 \tab 3 \tab 5 \tab 3 \tab 3 \tab 66 \tab 9 \tab 8 \tab 15 \tab 4 \tab 3 \tab 1 \tab 3 \tab 6 \tab 9 \tab 56 \tab 5 \tab 6 \tab 16 \tab 2 \tab 13 \tab 40 \tab - \tab 1 \tab 2 \tab 208 \tab 4 \tab 3 \tab 2 \tab 3 \tab 11 \tab 2 \tab 6 \tab 24 \tab 32 \tab 8 \tab 5 \tab 4 \tab 20 \tab 10 \tab 11 \tab 5 \tab 2 \tab 3 \tab - \tab 7 \tab 2 \tab - \tab 2 \tab 4 \tab 3 \tab 34 \tab 2 \tab - \tab 6 \tab 7 \tab 10 \tab 10 \tab - \tab 33 \tab 14 \tab 7 \tab 11 \tab 4 \tab - \tab 49 \tab 7 \tab 11 \tab 3 \tab 9 \tab 63 \tab 8 \tab 5 \tab 2 \tab 10 \tab 6 \tab 5 \tab 4 \tab - \tab 15 \tab 10 \tab - \tab 2 \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 8 \tab 8 \tab 8 \tab 12 \tab 16 \tab 156 \tab 28 \tab 6 \tab - \tab 4 \tab 27 \tab 4 \tab 1 \tab 6 \tab 4 \tab 6 \tab 12 \tab - \tab 6 \tab 3 \tab 22 \tab 2 \tab 1 \tab 1 \tab 2 \tab 30 \tab 18 \tab - \tab 17 \tab 3 \tab 3 \tab 2 \tab 5 \tab 3 \tab 2 \tab 16 \tab 16 \tab 15 \tab 18 \tab 10 \tab - \tab - \tab 1 \tab 15 \tab 1 \tab - \tab 6 \tab 1 \tab 1 \tab 2 \tab 19 \tab 6 \tab 4 \tab 5 \tab 20 \tab 1 \tab - \tab - \tab 1 \tab - \tab 6 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab 4 \tab 4 \tab 5 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 9 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 5 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab - \tab 8 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 452 \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab 7 \tab 1 \tab - \tab 2 \tab 4 \tab 2 \tab 23 \tab 1 \tab 11 \tab 7 \tab 7 \tab 59 \tab 1 \tab 2 \tab 31 \tab 36 \tab 49 \tab 7 \tab 12 \tab 9 \tab 1 \tab 1 \tab 7 \tab 6 \tab - \tab 3 \tab - \tab 3 \tab 9 \tab 13 \tab 45 \tab 2 \tab - \tab 1 \tab 1 \tab 50 \tab 8 \tab 8 \tab 14 \tab 4 \tab 1 \tab - \tab 4 \tab 2 \tab 3 \tab 31 \tab 10 \tab 13 \tab 11 \tab - \tab 14 \tab 13 \tab 1 \tab 1 \tab 3 \tab 144 \tab 3 \tab 1 \tab 3 \tab 4 \tab 8 \tab 4 \tab 7 \tab 20 \tab 35 \tab 7 \tab 2 \tab 2 \tab 11 \tab 1 \tab 11 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 27 \tab 1 \tab 2 \tab 2 \tab 9 \tab 6 \tab 13 \tab - \tab 24 \tab 6 \tab 8 \tab 3 \tab 3 \tab 1 \tab 24 \tab 3 \tab 4 \tab 1 \tab 7 \tab 37 \tab 2 \tab - \tab 3 \tab 10 \tab 4 \tab 2 \tab 3 \tab 3 \tab 7 \tab 3 \tab - \tab - \tab 2 \tab 6 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 5 \tab 2 \tab 4 \tab 7 \tab 20 \tab 95 \tab 16 \tab 11 \tab 1 \tab 1 \tab 21 \tab 3 \tab - \tab 4 \tab - \tab 3 \tab 11 \tab - \tab 12 \tab 4 \tab 8 \tab 5 \tab - \tab 1 \tab 1 \tab 21 \tab 6 \tab - \tab 11 \tab 4 \tab 3 \tab - \tab 7 \tab 5 \tab 4 \tab 9 \tab 19 \tab 11 \tab 18 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 9 \tab 2 \tab 6 \tab 2 \tab 9 \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 6 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab 5 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 4 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 466 \tab 2 \tab - \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 5 \tab 2 \tab 1 \tab 1 \tab 4 \tab 4 \tab 29 \tab 1 \tab 12 \tab 7 \tab 8 \tab 33 \tab 1 \tab 3 \tab 30 \tab 34 \tab 32 \tab 5 \tab 21 \tab 4 \tab 2 \tab - \tab 10 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab 7 \tab 14 \tab 60 \tab 1 \tab - \tab 8 \tab 4 \tab 52 \tab 8 \tab 9 \tab 7 \tab 11 \tab 2 \tab - \tab - \tab 5 \tab 5 \tab 36 \tab 8 \tab 12 \tab 10 \tab - \tab 13 \tab 23 \tab 1 \tab - \tab 4 \tab 124 \tab 4 \tab 1 \tab 2 \tab - \tab 8 \tab 4 \tab 5 \tab 28 \tab 18 \tab 5 \tab 1 \tab 2 \tab 28 \tab 3 \tab 8 \tab 1 \tab 1 \tab 4 \tab 2 \tab 4 \tab 4 \tab - \tab - \tab 2 \tab - \tab 27 \tab 1 \tab 1 \tab 1 \tab 8 \tab 5 \tab 5 \tab - \tab 38 \tab 8 \tab 2 \tab 5 \tab 1 \tab - \tab 27 \tab 13 \tab 3 \tab 1 \tab 2 \tab 38 \tab 2 \tab 1 \tab 4 \tab 3 \tab 2 \tab - \tab 2 \tab 7 \tab 5 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 7 \tab 4 \tab 2 \tab 4 \tab 11 \tab 16 \tab 138 \tab 15 \tab 5 \tab - \tab 2 \tab 11 \tab 10 \tab - \tab 7 \tab - \tab 5 \tab 5 \tab 2 \tab 11 \tab 4 \tab 6 \tab - \tab 2 \tab - \tab - \tab 16 \tab 8 \tab - \tab 14 \tab 3 \tab 2 \tab 1 \tab 8 \tab 1 \tab 1 \tab 6 \tab 13 \tab 14 \tab 16 \tab 3 \tab 1 \tab - \tab - \tab 7 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 2 \tab 9 \tab 3 \tab 6 \tab 5 \tab 21 \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab 3 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 4 \tab 4 \tab - \tab 2 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 47296 \tab 33 \tab 15 \tab 21 \tab 45 \tab 3 \tab 31 \tab 35 \tab 74 \tab 32 \tab 12 \tab 39 \tab 42 \tab 39 \tab 452 \tab 41 \tab 270 \tab 108 \tab 105 \tab 383 \tab 17 \tab 35 \tab 353 \tab 563 \tab 749 \tab 69 \tab 307 \tab 101 \tab 30 \tab 5 \tab 146 \tab 68 \tab 28 \tab 32 \tab 7 \tab 24 \tab 179 \tab 162 \tab 863 \tab 18 \tab 22 \tab 87 \tab 32 \tab 678 \tab 112 \tab 61 \tab 153 \tab 87 \tab 32 \tab 8 \tab 25 \tab 70 \tab 83 \tab 596 \tab 91 \tab 135 \tab 139 \tab 24 \tab 221 \tab 452 \tab 15 \tab 5 \tab 38 \tab 2957 \tab 54 \tab 16 \tab 31 \tab 58 \tab 136 \tab 24 \tab 93 \tab 284 \tab 375 \tab 132 \tab 60 \tab 30 \tab 164 \tab 68 \tab 126 \tab 27 \tab 19 \tab 23 \tab 13 \tab 88 \tab 30 \tab 2 \tab 11 \tab 37 \tab 23 \tab 314 \tab 21 \tab 71 \tab 49 \tab 84 \tab 106 \tab 145 \tab 16 \tab 525 \tab 115 \tab 56 \tab 86 \tab 29 \tab 7 \tab 276 \tab 79 \tab 92 \tab 27 \tab 93 \tab 847 \tab 50 \tab 25 \tab 36 \tab 108 \tab 56 \tab 25 \tab 30 \tab 49 \tab 159 \tab 68 \tab 6 \tab 18 \tab 18 \tab 30 \tab 19 \tab 15 \tab 19 \tab 19 \tab 51 \tab 59 \tab 60 \tab 55 \tab 120 \tab 269 \tab 1889 \tab 365 \tab 98 \tab 3 \tab 13 \tab 215 \tab 63 \tab 6 \tab 51 \tab 14 \tab 43 \tab 147 \tab 6 \tab 124 \tab 67 \tab 197 \tab 32 \tab 22 \tab 17 \tab 23 \tab 265 \tab 129 \tab 3 \tab 185 \tab 23 \tab 28 \tab 21 \tab 67 \tab 45 \tab 19 \tab 148 \tab 164 \tab 161 \tab 213 \tab 70 \tab 6 \tab 5 \tab 15 \tab 85 \tab 3 \tab 2 \tab 43 \tab 13 \tab 7 \tab 14 \tab 168 \tab 54 \tab 71 \tab 51 \tab 179 \tab 10 \tab 1 \tab 12 \tab 21 \tab 4 \tab 25 \tab 19 \tab 10 \tab 2 \tab 9 \tab 4 \tab 8 \tab 27 \tab 26 \tab 11 \tab 5 \tab 4 \tab 12 \tab 23 \tab 13 \tab 22 \tab 1 \tab 18 \tab 11 \tab 13 \tab 41 \tab 21 \tab 4 \tab 18 \tab 17 \tab 15 \tab 5 \tab 14 \tab 25 \tab 24 \tab 15 \tab 10 \tab 3 \tab 47 \tab 32 \tab 12 \tab 19 \tab 16 \tab 1 \tab 17 \tab 35 \tab 65 \tab 31 \tab 5 \tab 12 \tab 5 \tab 3 \tab 10 \tab 38 \tab 7 \tab 14 \tab 3 \tab 21 \tab 4 \tab 4 \tab 10 \tab 6 \tab 10 \tab 29 \tab 30 \tab 7 \tab 6 \tab 15 \tab 28 \tab 11 \tab 2 \tab 5 \tab 1 \tab 21 \tab 2 \tab 7 \tab 4 \tab 3 \tab 6 \tab 10 \tab 13 \tab 11 \tab 3 \tab 6 \tab 13 \tab 5 \tab 8 \tab 17 \tab 12 \tab 2 \tab 3 \tab 61 \tab 9 \tab 1 \tab 7 \tab 6 \tab 4 \tab 6 \tab 2 \tab 11 \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab 2 \tab 5 \tab 6 \tab 6 \tab 1 \tab 2 \tab 5 \tab 4 \tab 1 \tab 5 \tab 2 \tab 3 \tab 1 \tab 2 \tab 2 \tab 6 \tab 17 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 3 \tab 7 \tab 3 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} i don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
