% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_597.R
\name{pasei10}
\alias{pasei10}
\title{pasei10}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's father's socioeconomic index (2010)
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 11.8 \tab 12.4 \tab 12.6 \tab 13.2 \tab 13.6 \tab 14 \tab 14.4 \tab 14.8 \tab 15.8 \tab 17.1 \tab 18.6 \tab 18.8 \tab 19.7 \tab 20 \tab 20.3 \tab 20.5 \tab 20.7 \tab 20.8 \tab 20.9 \tab 21 \tab 21.4 \tab 21.5 \tab 21.6 \tab 22.9 \tab 23 \tab 23.2 \tab 23.3 \tab 23.7 \tab 23.8 \tab 24 \tab 24.1 \tab 24.2 \tab 24.6 \tab 25 \tab 25.1 \tab 25.2 \tab 25.4 \tab 26.1 \tab 26.2 \tab 26.5 \tab 26.6 \tab 26.8 \tab 26.9 \tab 27.5 \tab 27.7 \tab 28.3 \tab 28.4 \tab 28.5 \tab 28.6 \tab 28.7 \tab 29.5 \tab 29.9 \tab 30.1 \tab 30.8 \tab 31 \tab 31.1 \tab 31.2 \tab 31.3 \tab 31.5 \tab 31.6 \tab 31.9 \tab 32 \tab 32.6 \tab 33.1 \tab 33.2 \tab 33.3 \tab 33.4 \tab 33.7 \tab 33.9 \tab 34.1 \tab 34.6 \tab 34.7 \tab 35.4 \tab 35.8 \tab 36.2 \tab 36.4 \tab 36.7 \tab 36.8 \tab 37 \tab 37.1 \tab 38.1 \tab 38.2 \tab 38.3 \tab 38.8 \tab 39.1 \tab 39.2 \tab 39.7 \tab 39.9 \tab 40 \tab 40.1 \tab 41 \tab 41.1 \tab 41.4 \tab 41.8 \tab 42.2 \tab 42.8 \tab 43 \tab 45.3 \tab 45.4 \tab 46 \tab 46.5 \tab 46.6 \tab 47.9 \tab 48.3 \tab 48.8 \tab 49.4 \tab 49.6 \tab 50 \tab 50.1 \tab 50.3 \tab 50.4 \tab 50.7 \tab 51.5 \tab 51.8 \tab 51.9 \tab 52 \tab 52.1 \tab 52.3 \tab 52.5 \tab 53.5 \tab 54.6 \tab 55.3 \tab 56.9 \tab 57.1 \tab 57.4 \tab 57.5 \tab 57.8 \tab 58.1 \tab 58.4 \tab 59.1 \tab 59.8 \tab 60.3 \tab 60.4 \tab 61.4 \tab 62 \tab 62.6 \tab 62.7 \tab 63.1 \tab 63.3 \tab 64.2 \tab 65.1 \tab 65.3 \tab 66.1 \tab 67.7 \tab 68 \tab 68.6 \tab 69.3 \tab 69.4 \tab 71 \tab 72.6 \tab 73.5 \tab 73.9 \tab 75.2 \tab 75.5 \tab 76.3 \tab 77.4 \tab 77.9 \tab 79.4 \tab 80.1 \tab 81.6 \tab 82.4 \tab 82.5 \tab 83.1 \tab 83.4 \tab 83.8 \tab 84 \tab 84.3 \tab 84.5 \tab 85.2 \tab 86.5 \tab 86.6 \tab 86.7 \tab 86.8 \tab 87.6 \tab 89.3 \tab 9 \tab 91.1 \tab 91.9 \tab 92 \tab 92.1 \tab 92.8 \tab iap \tab 12.7 \tab 17.5 \tab 19.2 \tab 21.1 \tab 22.1 \tab 24.8 \tab 25.7 \tab 27 \tab 27.1 \tab 27.2 \tab 27.4 \tab 28.8 \tab 29.6 \tab 32.7 \tab 34.4 \tab 35.1 \tab 35.2 \tab 35.3 \tab 37.2 \tab 37.3 \tab 43.1 \tab 45.1 \tab 48.1 \tab 48.4 \tab 52.4 \tab 53.6 \tab 53.8 \tab 53.9 \tab 54.3 \tab 58.2 \tab 59 \tab 59.4 \tab 59.9 \tab 60.1 \tab 61.5 \tab 62.1 \tab 65.2 \tab 65.6 \tab 66.4 \tab 67.8 \tab 70.1 \tab 72.5 \tab 72.7 \tab 73.1 \tab 74.6 \tab 77.3 \tab 80.9 \tab 89.1 \tab 89.9 \tab 91.6 \tab 10.6 \tab 19.6 \tab 21.3 \tab 22.5 \tab 27.6 \tab 29.7 \tab 32.1 \tab 34.8 \tab 37.7 \tab 39.8 \tab 41.6 \tab 43.4 \tab 43.5 \tab 44.9 \tab 46.2 \tab 51.6 \tab 52.9 \tab 57 \tab 61.2 \tab 62.9 \tab 63.7 \tab 70.3 \tab 81 \tab 85 \tab 86.1 \tab 87.8 \tab 25.5 \tab 37.6 \tab 39.6 \tab 42.3 \tab 53.2 \tab 53.7 \tab 57.3 \tab 71.8 \tab 73.4 \tab 74.8 \tab 77.7 \tab 78.3 \tab 80.7 \tab 13.3 \tab 14.6 \tab 18.9 \tab 21.8 \tab 29.4 \tab 37.4 \tab 44.5 \tab 49.3 \tab 71.6 \tab 73.6 \tab 82.7 \tab 85.5 \tab 21.2 \tab 28.1 \tab 42.7 \tab 55.1 \tab 63.2 \tab 30.9 \tab 40.4 \tab 44.6 \tab 53.1 \tab 60.2 \tab 61.9 \tab 64.9 \tab 66.2 \tab 84.2 \tab 93.7 \tab 25.6 \tab 34.5 \tab 56.8 \tab 60.5 \tab 62.4 \tab 79.3 \tab 30.3 \tab 42.9 \tab 45 \tab 51.7 \tab 31.4 \tab 39.5 \tab 53 \tab 71.7 \tab 80.2 \tab 82.9 \tab 85.8 \tab 87.9 \tab 88.7 \tab 87.3 \tab 88 \tab 19 \tab 81.9 \tab 88.3 \tab 29.2 \tab 73.2 \tab 88.6 \tab 35.5 \tab 39 \tab 81.5 \tab 85.9 \tab 26.7 \tab 32.8 \tab 76.5 \tab 55 \tab 64.4 \tab 20.1 \tab 43.8 \tab 54.1 \tab 70.4 \tab 73 \tab 75.6 \tab 77 \tab 79.9 \tab 86.9 \tab 38.4 \tab 75.7 \tab 41.2 \tab 49.2 \tab 64.5 \tab 70.5 \tab 71.1 \tab 46.7 \tab 29.3 \tab 77.2 \tab 22.4 \tab 30 \tab 37.9 \tab 37.8 \tab 17 \tab 69.5 \tab Total \cr
   1972 \tab 3 \tab 2 \tab 2 \tab 2 \tab 6 \tab 56 \tab 3 \tab 4 \tab 2 \tab 2 \tab 7 \tab 7 \tab 8 \tab 13 \tab 2 \tab 1 \tab 17 \tab 2 \tab 7 \tab 3 \tab 3 \tab 12 \tab 1 \tab 2 \tab 1 \tab 5 \tab 22 \tab 5 \tab 4 \tab 11 \tab 1 \tab 1 \tab 9 \tab 7 \tab 3 \tab 14 \tab 1 \tab 2 \tab 2 \tab 8 \tab 4 \tab 7 \tab 3 \tab 1 \tab 3 \tab 4 \tab 5 \tab 6 \tab 1 \tab 34 \tab 1 \tab 7 \tab 2 \tab 1 \tab 7 \tab 2 \tab 17 \tab 2 \tab 1 \tab 3 \tab 2 \tab 38 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab 1 \tab 3 \tab 2 \tab 8 \tab 7 \tab 4 \tab 15 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 12 \tab 2 \tab 17 \tab 5 \tab 1 \tab 3 \tab 9 \tab 301 \tab 1 \tab 3 \tab 5 \tab 1 \tab 1 \tab 2 \tab 31 \tab 4 \tab 15 \tab 1 \tab 5 \tab 11 \tab 25 \tab 5 \tab 1 \tab 5 \tab 6 \tab 2 \tab 2 \tab 17 \tab 2 \tab 3 \tab 7 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 5 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 1 \tab 4 \tab 9 \tab 5 \tab 4 \tab 2 \tab 7 \tab 1 \tab 6 \tab 2 \tab 12 \tab 11 \tab 1 \tab 1 \tab 1 \tab 1 \tab 7 \tab 12 \tab 1 \tab 16 \tab 4 \tab 4 \tab 7 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 9 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 1 \tab 7 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 5 \tab 1 \tab 3 \tab 3 \tab 8 \tab 406 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1 \tab - \tab - \tab 2 \tab 5 \tab 81 \tab 5 \tab 2 \tab - \tab 1 \tab 6 \tab 5 \tab 2 \tab 7 \tab - \tab 1 \tab 9 \tab 3 \tab 5 \tab 2 \tab 1 \tab 6 \tab 1 \tab 4 \tab - \tab - \tab 22 \tab 1 \tab 2 \tab 12 \tab - \tab 1 \tab 11 \tab 8 \tab 5 \tab 18 \tab - \tab 3 \tab 1 \tab 4 \tab 2 \tab 26 \tab 5 \tab 1 \tab 5 \tab 1 \tab 7 \tab 7 \tab 1 \tab 45 \tab 1 \tab 1 \tab 2 \tab 2 \tab 9 \tab 9 \tab 10 \tab - \tab 1 \tab 2 \tab 3 \tab 38 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 9 \tab 3 \tab 2 \tab 6 \tab 6 \tab 4 \tab 8 \tab 6 \tab 7 \tab - \tab 2 \tab - \tab 8 \tab 1 \tab 20 \tab 5 \tab - \tab 3 \tab 16 \tab 233 \tab 2 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 35 \tab 1 \tab 35 \tab 1 \tab 9 \tab 23 \tab 33 \tab 10 \tab 1 \tab 4 \tab 3 \tab 1 \tab 1 \tab 19 \tab 3 \tab 2 \tab 3 \tab - \tab 2 \tab - \tab 3 \tab 6 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab 6 \tab 9 \tab 1 \tab 5 \tab 7 \tab - \tab 1 \tab 4 \tab 2 \tab 21 \tab 13 \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab 9 \tab - \tab 16 \tab 6 \tab 5 \tab 4 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 3 \tab 11 \tab 7 \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 6 \tab 3 \tab 1 \tab 3 \tab 5 \tab 4 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 7 \tab 2 \tab 4 \tab 254 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab - \tab 4 \tab 4 \tab 4 \tab 63 \tab 3 \tab - \tab 1 \tab - \tab 8 \tab 8 \tab 5 \tab 4 \tab - \tab 1 \tab 10 \tab - \tab 4 \tab 1 \tab - \tab 9 \tab 1 \tab 8 \tab 1 \tab 2 \tab 15 \tab 2 \tab 6 \tab 8 \tab 2 \tab 3 \tab 12 \tab 2 \tab 4 \tab 13 \tab 2 \tab 6 \tab 3 \tab 3 \tab - \tab 19 \tab 2 \tab 1 \tab 3 \tab 3 \tab 7 \tab 3 \tab - \tab 47 \tab 1 \tab 2 \tab - \tab - \tab 7 \tab 6 \tab 16 \tab 2 \tab - \tab 5 \tab 2 \tab 33 \tab 8 \tab - \tab - \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 4 \tab 9 \tab 4 \tab 3 \tab 18 \tab 5 \tab 11 \tab 2 \tab - \tab - \tab 9 \tab 2 \tab 11 \tab 10 \tab 1 \tab 8 \tab 16 \tab 231 \tab 3 \tab 1 \tab 9 \tab - \tab - \tab 5 \tab 29 \tab 3 \tab 26 \tab - \tab 4 \tab 13 \tab 22 \tab 6 \tab 1 \tab - \tab 4 \tab 2 \tab - \tab 18 \tab 2 \tab - \tab 8 \tab - \tab - \tab - \tab 1 \tab 4 \tab 4 \tab 4 \tab 2 \tab - \tab 1 \tab 1 \tab 5 \tab - \tab 2 \tab 4 \tab 8 \tab 2 \tab 8 \tab 5 \tab 7 \tab 1 \tab 4 \tab 2 \tab 19 \tab 7 \tab 1 \tab 4 \tab - \tab - \tab - \tab 9 \tab - \tab 15 \tab 2 \tab 4 \tab 4 \tab 1 \tab 4 \tab - \tab 1 \tab 7 \tab - \tab 2 \tab 11 \tab 8 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab 2 \tab 4 \tab 5 \tab 5 \tab 4 \tab - \tab - \tab 4 \tab 2 \tab - \tab 3 \tab 2 \tab 4 \tab 5 \tab 3 \tab 294 \tab 2 \tab 2 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1 \tab - \tab 1 \tab 6 \tab 1 \tab 25 \tab 6 \tab 2 \tab - \tab 2 \tab 3 \tab 10 \tab 1 \tab 10 \tab - \tab 5 \tab 11 \tab - \tab 4 \tab - \tab - \tab 10 \tab - \tab 1 \tab 3 \tab 4 \tab 18 \tab 2 \tab 1 \tab 12 \tab 3 \tab - \tab 8 \tab 5 \tab 3 \tab 18 \tab - \tab 9 \tab 2 \tab 5 \tab 2 \tab 14 \tab 3 \tab 1 \tab 2 \tab 1 \tab 12 \tab 4 \tab - \tab 46 \tab 2 \tab 1 \tab - \tab 4 \tab 7 \tab 5 \tab 20 \tab 3 \tab 1 \tab 6 \tab 2 \tab 56 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 11 \tab 6 \tab 1 \tab 14 \tab 3 \tab 10 \tab 1 \tab - \tab - \tab 10 \tab - \tab 20 \tab 5 \tab 2 \tab 1 \tab 14 \tab 261 \tab 4 \tab 1 \tab 8 \tab - \tab - \tab 3 \tab 28 \tab - \tab 27 \tab 1 \tab 8 \tab 17 \tab 26 \tab 9 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 15 \tab 2 \tab 1 \tab 8 \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab 9 \tab - \tab - \tab 1 \tab 1 \tab 10 \tab - \tab 1 \tab 8 \tab 6 \tab 5 \tab 4 \tab 6 \tab 9 \tab - \tab 6 \tab - \tab 7 \tab 8 \tab 1 \tab 2 \tab - \tab - \tab 4 \tab 10 \tab - \tab 22 \tab 5 \tab 3 \tab 9 \tab - \tab - \tab 2 \tab 4 \tab 2 \tab - \tab 2 \tab 7 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 6 \tab 1 \tab 1 \tab 3 \tab 2 \tab 5 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 5 \tab 2 \tab 7 \tab 289 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 3 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab 4 \tab 4 \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 1 \tab 1 \tab 3 \tab 3 \tab 42 \tab 2 \tab 1 \tab 3 \tab 2 \tab 6 \tab 3 \tab 3 \tab 9 \tab 1 \tab 2 \tab 16 \tab 1 \tab 4 \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab - \tab 4 \tab 32 \tab 1 \tab 5 \tab 12 \tab 2 \tab 1 \tab 8 \tab 6 \tab 8 \tab 24 \tab 1 \tab 6 \tab 6 \tab 5 \tab 1 \tab 24 \tab 2 \tab 2 \tab 9 \tab 1 \tab 5 \tab 2 \tab - \tab 36 \tab 1 \tab - \tab - \tab 1 \tab 9 \tab 3 \tab 19 \tab 5 \tab - \tab 3 \tab - \tab 47 \tab 12 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 7 \tab 1 \tab 3 \tab 7 \tab 4 \tab 4 \tab 4 \tab 1 \tab 7 \tab - \tab 1 \tab - \tab 7 \tab - \tab 23 \tab 9 \tab - \tab 7 \tab 8 \tab 250 \tab 6 \tab - \tab 5 \tab 2 \tab - \tab 2 \tab 29 \tab 6 \tab 39 \tab - \tab 8 \tab 21 \tab 30 \tab 11 \tab - \tab 3 \tab 2 \tab 2 \tab 1 \tab 15 \tab 4 \tab - \tab 5 \tab - \tab 2 \tab 1 \tab - \tab 7 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 9 \tab - \tab 2 \tab 8 \tab 11 \tab 8 \tab 2 \tab 4 \tab 5 \tab - \tab 4 \tab 1 \tab 16 \tab 11 \tab 3 \tab 4 \tab - \tab - \tab 1 \tab 12 \tab - \tab 15 \tab 3 \tab 5 \tab 4 \tab - \tab - \tab - \tab - \tab 6 \tab 3 \tab 2 \tab 7 \tab 8 \tab - \tab - \tab - \tab 3 \tab - \tab 5 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 3 \tab 3 \tab 6 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab 2 \tab 10 \tab 262 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1 \tab - \tab 1 \tab 3 \tab 4 \tab 43 \tab 3 \tab 1 \tab 3 \tab 4 \tab 9 \tab 3 \tab 8 \tab 11 \tab 3 \tab 3 \tab 13 \tab 1 \tab 2 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab 28 \tab 6 \tab 2 \tab 11 \tab 1 \tab - \tab 7 \tab 4 \tab 5 \tab 17 \tab 1 \tab 6 \tab 6 \tab 5 \tab 1 \tab 14 \tab 4 \tab 1 \tab 1 \tab 3 \tab 18 \tab - \tab - \tab 31 \tab 1 \tab - \tab - \tab 2 \tab 7 \tab 4 \tab 16 \tab 4 \tab - \tab 2 \tab 1 \tab 36 \tab 4 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 4 \tab 2 \tab 7 \tab 10 \tab 3 \tab 8 \tab 2 \tab 14 \tab - \tab 2 \tab 1 \tab 5 \tab - \tab 23 \tab 2 \tab 2 \tab 4 \tab 11 \tab 250 \tab 1 \tab 3 \tab 6 \tab - \tab - \tab 7 \tab 48 \tab 5 \tab 13 \tab - \tab 11 \tab 8 \tab 32 \tab 5 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 13 \tab 1 \tab 2 \tab 13 \tab - \tab - \tab - \tab 2 \tab 4 \tab 3 \tab 6 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 8 \tab 7 \tab 2 \tab 8 \tab 6 \tab 6 \tab - \tab 5 \tab 3 \tab 4 \tab 8 \tab 5 \tab 4 \tab - \tab - \tab 2 \tab 13 \tab - \tab 9 \tab 1 \tab 3 \tab 6 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 7 \tab 4 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 8 \tab - \tab 1 \tab 3 \tab 4 \tab 365 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 6 \tab 4 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 42 \tab 3 \tab 1 \tab 1 \tab 5 \tab 4 \tab 3 \tab 7 \tab 16 \tab 1 \tab 4 \tab 13 \tab 2 \tab 4 \tab - \tab 2 \tab 4 \tab 1 \tab 2 \tab - \tab 4 \tab 20 \tab 2 \tab 4 \tab 9 \tab 1 \tab 2 \tab 12 \tab 11 \tab 4 \tab 11 \tab 1 \tab 7 \tab 4 \tab 5 \tab 3 \tab 26 \tab 5 \tab 2 \tab 4 \tab 5 \tab 11 \tab 1 \tab - \tab 56 \tab 2 \tab 2 \tab 1 \tab 1 \tab 10 \tab 8 \tab 11 \tab 4 \tab 1 \tab 1 \tab 3 \tab 42 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 9 \tab 1 \tab 1 \tab 2 \tab 6 \tab 2 \tab 8 \tab 4 \tab 15 \tab 2 \tab - \tab - \tab 9 \tab 2 \tab 25 \tab 6 \tab 4 \tab 5 \tab 14 \tab 216 \tab 2 \tab 1 \tab 6 \tab - \tab - \tab 5 \tab 40 \tab 11 \tab 36 \tab 1 \tab 13 \tab 22 \tab 29 \tab 3 \tab - \tab 10 \tab 1 \tab - \tab - \tab 11 \tab 2 \tab 1 \tab 7 \tab - \tab - \tab 1 \tab 2 \tab 6 \tab 1 \tab 3 \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 11 \tab 8 \tab 4 \tab 7 \tab 6 \tab 4 \tab 2 \tab 7 \tab 3 \tab 24 \tab 6 \tab 4 \tab 8 \tab - \tab - \tab 7 \tab 16 \tab 1 \tab 17 \tab 11 \tab 3 \tab 4 \tab - \tab 1 \tab - \tab 3 \tab 4 \tab 2 \tab 1 \tab 10 \tab 5 \tab 1 \tab - \tab - \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 7 \tab - \tab - \tab 2 \tab 4 \tab 6 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 9 \tab - \tab 3 \tab - \tab 7 \tab 220 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab 4 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 11 \tab 3 \tab - \tab 1 \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 35 \tab 7 \tab 3 \tab 1 \tab 1 \tab 6 \tab 7 \tab 2 \tab 4 \tab 1 \tab 3 \tab 13 \tab 1 \tab 6 \tab 1 \tab 1 \tab 9 \tab 2 \tab 2 \tab 2 \tab 4 \tab 19 \tab - \tab 6 \tab 8 \tab 2 \tab 1 \tab 8 \tab 6 \tab 1 \tab 14 \tab - \tab 5 \tab 1 \tab 3 \tab 2 \tab 13 \tab - \tab 2 \tab 1 \tab 2 \tab 11 \tab 5 \tab - \tab 38 \tab - \tab 1 \tab - \tab 1 \tab 5 \tab 7 \tab 13 \tab 3 \tab 1 \tab 3 \tab 2 \tab 36 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 5 \tab 3 \tab 5 \tab 6 \tab 2 \tab 2 \tab 14 \tab 5 \tab 9 \tab 1 \tab - \tab - \tab 11 \tab 2 \tab 17 \tab 11 \tab 2 \tab 8 \tab 14 \tab 203 \tab 3 \tab - \tab 7 \tab 1 \tab - \tab 8 \tab 33 \tab 7 \tab 46 \tab - \tab 11 \tab 19 \tab 25 \tab 6 \tab - \tab 1 \tab 4 \tab - \tab - \tab 10 \tab - \tab 5 \tab 11 \tab - \tab 2 \tab 1 \tab 1 \tab 10 \tab 4 \tab - \tab - \tab - \tab 2 \tab 1 \tab 6 \tab - \tab 1 \tab 12 \tab 7 \tab 6 \tab 4 \tab 7 \tab 3 \tab 1 \tab 4 \tab 3 \tab 14 \tab 6 \tab 1 \tab 4 \tab - \tab - \tab 5 \tab 14 \tab 2 \tab 13 \tab 6 \tab 7 \tab 7 \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab 5 \tab 8 \tab 4 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab 4 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 9 \tab 1 \tab 5 \tab 3 \tab - \tab - \tab - \tab 3 \tab 2 \tab 11 \tab - \tab 2 \tab - \tab 9 \tab 282 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 6 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 6 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 2 \tab 1 \tab 3 \tab 3 \tab 58 \tab 5 \tab 1 \tab 2 \tab 1 \tab 8 \tab 10 \tab 7 \tab 12 \tab - \tab 1 \tab 27 \tab 2 \tab 4 \tab 3 \tab 4 \tab 7 \tab 1 \tab 1 \tab 2 \tab 4 \tab 35 \tab 3 \tab 7 \tab 10 \tab 4 \tab 2 \tab 18 \tab 9 \tab 3 \tab 24 \tab - \tab 4 \tab 4 \tab 9 \tab 3 \tab 29 \tab 4 \tab 1 \tab 5 \tab 4 \tab 9 \tab 4 \tab - \tab 46 \tab 3 \tab 2 \tab 5 \tab 4 \tab 10 \tab 2 \tab 18 \tab 1 \tab 1 \tab 10 \tab 2 \tab 64 \tab 2 \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 7 \tab 5 \tab 12 \tab 8 \tab 7 \tab 16 \tab 5 \tab 13 \tab 1 \tab 2 \tab 1 \tab 10 \tab - \tab 14 \tab 8 \tab 2 \tab 4 \tab 19 \tab 246 \tab 4 \tab 2 \tab 7 \tab - \tab - \tab 1 \tab 34 \tab 2 \tab 34 \tab - \tab 10 \tab 12 \tab 30 \tab 13 \tab - \tab 5 \tab 5 \tab 2 \tab - \tab 12 \tab 2 \tab 1 \tab 10 \tab - \tab 1 \tab 1 \tab 5 \tab 5 \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 8 \tab - \tab 4 \tab 7 \tab 10 \tab 1 \tab 11 \tab 16 \tab 4 \tab 2 \tab 7 \tab - \tab 30 \tab 12 \tab 4 \tab 3 \tab 1 \tab 1 \tab 9 \tab 8 \tab - \tab 18 \tab 8 \tab 2 \tab 5 \tab - \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 7 \tab 6 \tab 9 \tab 1 \tab 1 \tab 2 \tab 6 \tab - \tab 7 \tab - \tab - \tab 3 \tab 1 \tab - \tab 3 \tab 8 \tab 6 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 5 \tab - \tab 3 \tab 3 \tab 5 \tab 382 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 5 \tab 1 \tab - \tab - \tab 3 \tab - \tab 6 \tab 3 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 3 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 7 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 41 \tab 3 \tab 3 \tab 3 \tab 2 \tab 7 \tab 10 \tab 4 \tab 6 \tab 1 \tab 3 \tab 21 \tab 3 \tab 4 \tab 3 \tab 3 \tab 9 \tab 3 \tab 6 \tab 3 \tab 2 \tab 21 \tab 2 \tab 6 \tab 11 \tab 2 \tab - \tab 18 \tab 10 \tab 2 \tab 14 \tab - \tab 8 \tab 5 \tab 6 \tab 2 \tab 18 \tab 2 \tab 2 \tab 1 \tab - \tab 10 \tab 3 \tab - \tab 50 \tab 1 \tab 7 \tab 1 \tab 2 \tab 8 \tab 4 \tab 19 \tab 4 \tab 1 \tab 5 \tab 1 \tab 53 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 8 \tab 4 \tab 2 \tab 10 \tab 5 \tab 5 \tab 8 \tab 5 \tab 5 \tab - \tab 2 \tab - \tab 7 \tab - \tab 25 \tab 14 \tab 1 \tab 2 \tab 19 \tab 176 \tab 2 \tab 2 \tab 5 \tab - \tab 1 \tab 2 \tab 25 \tab 9 \tab 39 \tab 1 \tab 13 \tab 23 \tab 44 \tab 7 \tab - \tab 4 \tab 1 \tab 4 \tab 1 \tab 14 \tab 2 \tab 3 \tab 6 \tab - \tab - \tab - \tab 3 \tab 8 \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab 3 \tab 2 \tab 1 \tab 10 \tab 8 \tab 2 \tab 13 \tab 10 \tab 5 \tab - \tab 2 \tab 1 \tab 15 \tab 13 \tab 1 \tab 7 \tab - \tab - \tab 3 \tab 14 \tab - \tab 24 \tab 6 \tab 7 \tab 6 \tab 1 \tab 6 \tab 2 \tab 1 \tab 4 \tab 2 \tab 3 \tab 10 \tab 10 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 14 \tab 2 \tab 1 \tab 11 \tab 6 \tab 7 \tab 1 \tab - \tab 2 \tab - \tab 4 \tab - \tab 10 \tab - \tab 1 \tab 5 \tab 8 \tab 240 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 5 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab 5 \tab 7 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab - \tab - \tab 1 \tab 3 \tab 34 \tab 1 \tab - \tab - \tab 1 \tab 7 \tab 9 \tab 2 \tab 7 \tab 3 \tab - \tab 12 \tab 1 \tab 4 \tab - \tab 2 \tab 7 \tab 2 \tab 3 \tab 1 \tab 2 \tab 23 \tab 4 \tab 8 \tab 8 \tab 1 \tab 2 \tab 11 \tab 9 \tab 8 \tab 16 \tab 2 \tab 6 \tab 2 \tab 3 \tab 1 \tab 22 \tab 5 \tab 5 \tab - \tab 4 \tab 12 \tab 1 \tab - \tab 29 \tab - \tab 5 \tab 1 \tab 5 \tab 11 \tab 2 \tab 8 \tab 1 \tab 1 \tab 6 \tab 3 \tab 46 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 9 \tab 3 \tab 3 \tab 5 \tab 6 \tab 2 \tab 17 \tab 2 \tab 6 \tab - \tab 1 \tab 1 \tab 9 \tab - \tab 8 \tab 4 \tab 4 \tab 4 \tab 13 \tab 219 \tab 4 \tab 2 \tab 7 \tab - \tab 1 \tab - \tab 24 \tab 8 \tab 45 \tab 1 \tab 14 \tab 21 \tab 31 \tab 11 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 8 \tab 1 \tab - \tab 11 \tab - \tab 1 \tab - \tab 3 \tab 7 \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 6 \tab - \tab 4 \tab 10 \tab 10 \tab 4 \tab 5 \tab 11 \tab 3 \tab 2 \tab 3 \tab 3 \tab 19 \tab 6 \tab 4 \tab 6 \tab 1 \tab - \tab 4 \tab 5 \tab - \tab 22 \tab 10 \tab 4 \tab 6 \tab 1 \tab 2 \tab 4 \tab 4 \tab 3 \tab - \tab 3 \tab 10 \tab 13 \tab - \tab 1 \tab - \tab 4 \tab - \tab 3 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 5 \tab 4 \tab 10 \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 5 \tab 1 \tab 2 \tab 1 \tab 3 \tab 221 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab - \tab 5 \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 8 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 37 \tab 4 \tab 3 \tab 4 \tab 1 \tab 5 \tab 15 \tab 9 \tab 3 \tab 2 \tab 3 \tab 13 \tab 2 \tab 6 \tab - \tab 5 \tab 5 \tab 2 \tab 1 \tab 4 \tab 1 \tab 16 \tab 1 \tab 3 \tab 6 \tab 3 \tab 1 \tab 12 \tab 7 \tab 3 \tab 21 \tab 1 \tab 5 \tab 2 \tab 1 \tab - \tab 22 \tab 4 \tab 3 \tab 2 \tab 1 \tab 12 \tab 2 \tab - \tab 41 \tab - \tab 1 \tab 2 \tab 2 \tab 8 \tab 2 \tab 17 \tab 6 \tab - \tab - \tab 2 \tab 46 \tab 3 \tab - \tab 3 \tab 2 \tab 1 \tab 6 \tab 7 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 10 \tab 3 \tab 2 \tab - \tab - \tab - \tab 6 \tab - \tab 15 \tab 4 \tab - \tab 3 \tab 17 \tab 215 \tab 1 \tab 3 \tab 7 \tab - \tab - \tab 1 \tab 16 \tab 6 \tab 42 \tab - \tab 14 \tab 21 \tab 40 \tab 13 \tab - \tab 7 \tab 3 \tab 6 \tab - \tab 11 \tab 1 \tab 1 \tab 5 \tab 2 \tab 2 \tab - \tab 2 \tab 6 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 7 \tab 1 \tab 4 \tab 11 \tab 10 \tab 3 \tab 9 \tab 11 \tab 5 \tab 2 \tab 8 \tab 2 \tab 14 \tab 15 \tab - \tab 7 \tab - \tab 1 \tab 5 \tab 15 \tab - \tab 18 \tab 6 \tab 9 \tab 7 \tab - \tab 3 \tab - \tab 3 \tab 5 \tab 1 \tab 6 \tab 9 \tab 11 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab 3 \tab - \tab 8 \tab - \tab 1 \tab 5 \tab 7 \tab 4 \tab 4 \tab - \tab - \tab 5 \tab 3 \tab 3 \tab 8 \tab 1 \tab 3 \tab 3 \tab 10 \tab 248 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 8 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 4 \tab - \tab 3 \tab 1 \tab - \tab 3 \tab 4 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 42 \tab 3 \tab 3 \tab 2 \tab 2 \tab 8 \tab 3 \tab 2 \tab 2 \tab - \tab 1 \tab 13 \tab - \tab 10 \tab - \tab 4 \tab 4 \tab 3 \tab 2 \tab - \tab 3 \tab 18 \tab 1 \tab 7 \tab 8 \tab 1 \tab 1 \tab 4 \tab 5 \tab 2 \tab 15 \tab 1 \tab 3 \tab 3 \tab 7 \tab 5 \tab 23 \tab 5 \tab 3 \tab 4 \tab 2 \tab 14 \tab 4 \tab - \tab 38 \tab 1 \tab 2 \tab - \tab - \tab 6 \tab 3 \tab 11 \tab 3 \tab - \tab 4 \tab 3 \tab 35 \tab - \tab - \tab - \tab - \tab - \tab - \tab 8 \tab 4 \tab 3 \tab 8 \tab 5 \tab 3 \tab 13 \tab 4 \tab 6 \tab - \tab 1 \tab - \tab 8 \tab - \tab 23 \tab 9 \tab - \tab 3 \tab 26 \tab 193 \tab 2 \tab 2 \tab 6 \tab - \tab - \tab - \tab 13 \tab 6 \tab 39 \tab - \tab 13 \tab 16 \tab 32 \tab 8 \tab - \tab 8 \tab 4 \tab 1 \tab - \tab 11 \tab 4 \tab 1 \tab 5 \tab - \tab - \tab 2 \tab 4 \tab 7 \tab 2 \tab 5 \tab 1 \tab - \tab 5 \tab 1 \tab 6 \tab - \tab 5 \tab 5 \tab 8 \tab 3 \tab 6 \tab 12 \tab 7 \tab - \tab 4 \tab 7 \tab 26 \tab 9 \tab 2 \tab 7 \tab - \tab - \tab 4 \tab 11 \tab 1 \tab 27 \tab 9 \tab 7 \tab 5 \tab - \tab 1 \tab 3 \tab 1 \tab 8 \tab 1 \tab 3 \tab 4 \tab 10 \tab - \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 2 \tab - \tab 4 \tab 1 \tab - \tab 7 \tab 5 \tab 6 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab 8 \tab - \tab 2 \tab 3 \tab 5 \tab 253 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 6 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 2 \tab - \tab - \tab 1 \tab 6 \tab 65 \tab 1 \tab 1 \tab 4 \tab 2 \tab 5 \tab 9 \tab 5 \tab 5 \tab 1 \tab 4 \tab 15 \tab - \tab 5 \tab 1 \tab 4 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 21 \tab 3 \tab 6 \tab 13 \tab 2 \tab 2 \tab 12 \tab 7 \tab 5 \tab 22 \tab - \tab 3 \tab 4 \tab 3 \tab 6 \tab 33 \tab 1 \tab - \tab 2 \tab - \tab 20 \tab 6 \tab 1 \tab 51 \tab 3 \tab 3 \tab 2 \tab 2 \tab 6 \tab 1 \tab 13 \tab 4 \tab 1 \tab 8 \tab 2 \tab 64 \tab 4 \tab 1 \tab 4 \tab - \tab - \tab 3 \tab 5 \tab 3 \tab 2 \tab 9 \tab 4 \tab 2 \tab 12 \tab 5 \tab 5 \tab - \tab - \tab - \tab 6 \tab 2 \tab 12 \tab 11 \tab - \tab 2 \tab 12 \tab 251 \tab 8 \tab - \tab 6 \tab 1 \tab - \tab 2 \tab 28 \tab 7 \tab 30 \tab - \tab 10 \tab 21 \tab 33 \tab 7 \tab 1 \tab 5 \tab 2 \tab 3 \tab - \tab 8 \tab 2 \tab - \tab 12 \tab - \tab - \tab - \tab 2 \tab 7 \tab 3 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab - \tab 4 \tab 13 \tab 12 \tab 6 \tab 6 \tab 10 \tab 3 \tab 4 \tab 7 \tab - \tab 14 \tab 13 \tab 4 \tab 10 \tab - \tab - \tab 8 \tab 9 \tab - \tab 22 \tab 5 \tab 6 \tab 6 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 4 \tab 15 \tab 9 \tab - \tab 1 \tab - \tab 2 \tab - \tab 11 \tab 1 \tab - \tab 7 \tab 3 \tab - \tab 8 \tab 7 \tab 10 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 10 \tab - \tab 2 \tab 2 \tab 12 \tab 382 \tab 2 \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab 8 \tab - \tab 1 \tab - \tab 2 \tab - \tab 6 \tab 8 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 2 \tab - \tab - \tab - \tab 2 \tab 37 \tab 3 \tab 2 \tab 1 \tab 1 \tab 6 \tab 8 \tab 4 \tab 9 \tab - \tab 1 \tab 16 \tab 1 \tab 6 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 19 \tab 3 \tab 5 \tab 7 \tab 4 \tab 1 \tab 6 \tab 4 \tab 4 \tab 15 \tab - \tab 6 \tab 1 \tab 5 \tab 4 \tab 28 \tab 3 \tab 3 \tab 3 \tab 4 \tab 11 \tab 7 \tab - \tab 33 \tab 2 \tab 3 \tab 1 \tab 1 \tab 6 \tab 3 \tab 12 \tab - \tab 1 \tab 7 \tab 3 \tab 47 \tab 6 \tab - \tab 1 \tab - \tab - \tab 2 \tab 5 \tab 3 \tab 2 \tab 7 \tab 3 \tab 2 \tab 9 \tab 4 \tab 6 \tab - \tab 1 \tab - \tab 6 \tab 1 \tab 20 \tab 4 \tab 3 \tab 8 \tab 22 \tab 183 \tab 2 \tab 4 \tab 6 \tab 1 \tab - \tab 1 \tab 11 \tab 9 \tab 58 \tab - \tab 11 \tab 23 \tab 24 \tab 4 \tab - \tab 6 \tab 2 \tab 2 \tab - \tab 11 \tab 1 \tab 1 \tab 8 \tab - \tab - \tab 1 \tab - \tab 6 \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 4 \tab 9 \tab 5 \tab 3 \tab 6 \tab 11 \tab 3 \tab 2 \tab 4 \tab 1 \tab 21 \tab 11 \tab 5 \tab 6 \tab - \tab - \tab 2 \tab 13 \tab - \tab 22 \tab 10 \tab 9 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 5 \tab 1 \tab 7 \tab 10 \tab 14 \tab - \tab - \tab 1 \tab 1 \tab - \tab 6 \tab - \tab - \tab 8 \tab 5 \tab - \tab 7 \tab 5 \tab 7 \tab 3 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 7 \tab 1 \tab 2 \tab 2 \tab 8 \tab 227 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 9 \tab - \tab - \tab 1 \tab 1 \tab - \tab 7 \tab 7 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 5 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 5 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 2 \tab - \tab - \tab 5 \tab 3 \tab 45 \tab 2 \tab 1 \tab 5 \tab 4 \tab 7 \tab 5 \tab 3 \tab 5 \tab 1 \tab 2 \tab 7 \tab - \tab 4 \tab - \tab 1 \tab 6 \tab 1 \tab 5 \tab 3 \tab 1 \tab 22 \tab 4 \tab 4 \tab 6 \tab 1 \tab - \tab 13 \tab 11 \tab 8 \tab 21 \tab 3 \tab 4 \tab 1 \tab 2 \tab 1 \tab 24 \tab 3 \tab 1 \tab 1 \tab - \tab 8 \tab 2 \tab 1 \tab 24 \tab - \tab 2 \tab 1 \tab - \tab 7 \tab 6 \tab 21 \tab 3 \tab 1 \tab 5 \tab - \tab 65 \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab 8 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2 \tab 13 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 5 \tab - \tab 19 \tab 7 \tab 3 \tab 6 \tab 14 \tab 194 \tab - \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 21 \tab 6 \tab 32 \tab 1 \tab 18 \tab 20 \tab 36 \tab 7 \tab - \tab 9 \tab 4 \tab - \tab - \tab 12 \tab 1 \tab 3 \tab 11 \tab - \tab 1 \tab 3 \tab 2 \tab 5 \tab 2 \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab 13 \tab 1 \tab 4 \tab 13 \tab 8 \tab 3 \tab 9 \tab 10 \tab 9 \tab - \tab 7 \tab 2 \tab 16 \tab 12 \tab 2 \tab 2 \tab - \tab - \tab 7 \tab 13 \tab - \tab 24 \tab 10 \tab 7 \tab 8 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 7 \tab 12 \tab - \tab 2 \tab 1 \tab 5 \tab - \tab 9 \tab 2 \tab 1 \tab 10 \tab 2 \tab 1 \tab 11 \tab 8 \tab 4 \tab 3 \tab - \tab - \tab 1 \tab 3 \tab - \tab 7 \tab - \tab 4 \tab - \tab 6 \tab 247 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 10 \tab 2 \tab - \tab - \tab 1 \tab - \tab 4 \tab 7 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 10 \tab - \tab 1 \tab - \tab 1 \tab - \tab 6 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1 \tab - \tab - \tab - \tab 2 \tab 26 \tab 1 \tab - \tab 3 \tab - \tab 10 \tab 12 \tab 3 \tab 4 \tab - \tab 2 \tab 9 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab 23 \tab 3 \tab 4 \tab 11 \tab 2 \tab 3 \tab 6 \tab 9 \tab 9 \tab 20 \tab 1 \tab 3 \tab 4 \tab 3 \tab 4 \tab 26 \tab 1 \tab 1 \tab 2 \tab 2 \tab 19 \tab 1 \tab - \tab 33 \tab - \tab 1 \tab 1 \tab - \tab 7 \tab 2 \tab 14 \tab 2 \tab - \tab 5 \tab 1 \tab 45 \tab 5 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 8 \tab 2 \tab 3 \tab 6 \tab 2 \tab 6 \tab 12 \tab 5 \tab 2 \tab - \tab - \tab 1 \tab 7 \tab - \tab 11 \tab 8 \tab 5 \tab 3 \tab 19 \tab 150 \tab 1 \tab 3 \tab 5 \tab - \tab - \tab 3 \tab 14 \tab 12 \tab 44 \tab - \tab 5 \tab 8 \tab 23 \tab 8 \tab - \tab 12 \tab 2 \tab - \tab - \tab 8 \tab 2 \tab 1 \tab 7 \tab - \tab 1 \tab 2 \tab - \tab 5 \tab 3 \tab 1 \tab 5 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 3 \tab 24 \tab 7 \tab 4 \tab 6 \tab 8 \tab 4 \tab - \tab 3 \tab 3 \tab 17 \tab 8 \tab 2 \tab 3 \tab - \tab 1 \tab 3 \tab 8 \tab 1 \tab 20 \tab 11 \tab 10 \tab 5 \tab 3 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 7 \tab 10 \tab 1 \tab 1 \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 8 \tab - \tab 1 \tab 6 \tab 6 \tab 7 \tab 3 \tab - \tab 1 \tab - \tab 5 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 6 \tab 233 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 5 \tab 5 \tab 4 \tab 3 \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 40 \tab 6 \tab 1 \tab 1 \tab 3 \tab 10 \tab 4 \tab 3 \tab 3 \tab 1 \tab 1 \tab 19 \tab - \tab 3 \tab - \tab 3 \tab 6 \tab - \tab 2 \tab 2 \tab 1 \tab 22 \tab 1 \tab 2 \tab 12 \tab - \tab - \tab 10 \tab 12 \tab 1 \tab 18 \tab 1 \tab 5 \tab 4 \tab 4 \tab 1 \tab 21 \tab 4 \tab 1 \tab 4 \tab 2 \tab 11 \tab 2 \tab - \tab 33 \tab 2 \tab - \tab 1 \tab 1 \tab 9 \tab 6 \tab 15 \tab 2 \tab - \tab 5 \tab 3 \tab 51 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 9 \tab - \tab 3 \tab 7 \tab 1 \tab 8 \tab 14 \tab 3 \tab 2 \tab - \tab 2 \tab - \tab 4 \tab 1 \tab 12 \tab 11 \tab 4 \tab 3 \tab 12 \tab 173 \tab - \tab - \tab 9 \tab 1 \tab - \tab - \tab 17 \tab 6 \tab 53 \tab - \tab 14 \tab 20 \tab 24 \tab 11 \tab 1 \tab 8 \tab 3 \tab 1 \tab 2 \tab 7 \tab 2 \tab - \tab 6 \tab - \tab - \tab 1 \tab 3 \tab 13 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 5 \tab - \tab 4 \tab 11 \tab 7 \tab 2 \tab 11 \tab 14 \tab 6 \tab - \tab 2 \tab 4 \tab 29 \tab 8 \tab 3 \tab 7 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 25 \tab 7 \tab 3 \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab 7 \tab 12 \tab 6 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 7 \tab 1 \tab 2 \tab 10 \tab 2 \tab 1 \tab 3 \tab 6 \tab 11 \tab 4 \tab - \tab - \tab 2 \tab 3 \tab - \tab 3 \tab - \tab 2 \tab 4 \tab 10 \tab 266 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 6 \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab 6 \tab 1 \tab 2 \tab - \tab - \tab - \tab 10 \tab 7 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 10 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 6 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 3 \tab 2 \tab 3 \tab - \tab 42 \tab 3 \tab 2 \tab 2 \tab - \tab 9 \tab 6 \tab 3 \tab 9 \tab - \tab 3 \tab 17 \tab 2 \tab 2 \tab 1 \tab 2 \tab 3 \tab 3 \tab 2 \tab 3 \tab 3 \tab 22 \tab 1 \tab 3 \tab 7 \tab 1 \tab - \tab 9 \tab 11 \tab 3 \tab 10 \tab - \tab 5 \tab 3 \tab 1 \tab 2 \tab 25 \tab 1 \tab 3 \tab 4 \tab 4 \tab 11 \tab 4 \tab - \tab 36 \tab - \tab 5 \tab 1 \tab 2 \tab 10 \tab 6 \tab 21 \tab 1 \tab 1 \tab 3 \tab 2 \tab 49 \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 9 \tab 2 \tab 3 \tab 9 \tab 6 \tab 10 \tab 1 \tab 5 \tab - \tab - \tab - \tab 12 \tab 1 \tab 15 \tab 7 \tab - \tab 2 \tab 24 \tab 189 \tab 2 \tab 2 \tab 9 \tab - \tab 1 \tab 4 \tab 4 \tab 7 \tab 43 \tab 1 \tab 6 \tab 26 \tab 30 \tab 8 \tab - \tab 13 \tab 2 \tab 1 \tab - \tab 8 \tab 2 \tab - \tab 7 \tab 1 \tab - \tab 3 \tab 2 \tab 10 \tab 5 \tab 2 \tab 2 \tab 2 \tab 1 \tab 3 \tab 4 \tab - \tab 2 \tab 13 \tab 11 \tab 7 \tab 8 \tab 16 \tab 5 \tab 3 \tab 4 \tab 3 \tab 23 \tab 12 \tab - \tab 4 \tab - \tab - \tab 9 \tab 17 \tab - \tab 29 \tab 8 \tab 8 \tab 10 \tab 1 \tab 1 \tab - \tab - \tab 13 \tab - \tab 7 \tab 11 \tab 15 \tab - \tab 5 \tab - \tab 3 \tab - \tab 9 \tab - \tab - \tab 5 \tab 3 \tab 2 \tab 8 \tab 10 \tab 6 \tab 4 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab 11 \tab - \tab 2 \tab 4 \tab 13 \tab 244 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 10 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 5 \tab 9 \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab 2 \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 2 \tab - \tab 2 \tab 6 \tab 58 \tab 8 \tab 2 \tab - \tab 3 \tab 16 \tab 13 \tab 6 \tab 12 \tab 1 \tab 6 \tab 44 \tab - \tab 8 \tab - \tab 2 \tab 10 \tab 2 \tab 2 \tab - \tab 5 \tab 50 \tab 1 \tab 9 \tab 16 \tab 1 \tab 1 \tab 14 \tab 25 \tab 6 \tab 39 \tab 1 \tab 10 \tab 7 \tab 7 \tab 6 \tab 42 \tab 5 \tab 5 \tab 5 \tab 6 \tab 27 \tab 9 \tab - \tab 54 \tab 2 \tab 5 \tab 3 \tab 6 \tab 13 \tab 5 \tab 31 \tab 4 \tab - \tab 7 \tab 5 \tab 107 \tab 14 \tab - \tab 7 \tab 2 \tab 1 \tab 5 \tab 9 \tab 7 \tab 5 \tab 11 \tab 3 \tab 15 \tab 18 \tab 10 \tab 7 \tab 1 \tab 1 \tab - \tab 14 \tab 3 \tab 34 \tab 15 \tab 3 \tab 10 \tab 26 \tab 289 \tab 9 \tab - \tab 14 \tab - \tab - \tab 1 \tab 19 \tab 19 \tab 69 \tab 2 \tab 21 \tab 30 \tab 46 \tab 17 \tab 3 \tab 9 \tab 2 \tab 2 \tab - \tab 20 \tab 6 \tab 3 \tab 9 \tab - \tab 2 \tab 3 \tab 4 \tab 10 \tab 9 \tab 4 \tab 7 \tab 1 \tab 1 \tab 6 \tab 7 \tab 2 \tab 7 \tab 20 \tab 18 \tab 8 \tab 14 \tab 20 \tab 6 \tab 1 \tab 13 \tab 3 \tab 39 \tab 23 \tab 5 \tab 15 \tab 1 \tab - \tab 12 \tab 16 \tab 2 \tab 64 \tab 14 \tab 13 \tab 5 \tab 1 \tab 3 \tab 6 \tab 2 \tab 17 \tab 3 \tab 10 \tab 22 \tab 32 \tab 3 \tab 5 \tab 1 \tab 1 \tab - \tab 13 \tab 1 \tab 5 \tab 14 \tab 4 \tab 4 \tab 15 \tab 17 \tab 14 \tab 7 \tab - \tab 1 \tab 13 \tab 8 \tab - \tab 24 \tab - \tab 8 \tab 6 \tab 11 \tab 551 \tab 6 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab - \tab 1 \tab 4 \tab 3 \tab 2 \tab 1 \tab 2 \tab 13 \tab 2 \tab 4 \tab - \tab 1 \tab 1 \tab 8 \tab 13 \tab 9 \tab 3 \tab 4 \tab 1 \tab 4 \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab 4 \tab 2 \tab - \tab 1 \tab 1 \tab 11 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 9 \tab 1 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 5 \tab 1 \tab 2 \tab 5 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2 \tab - \tab - \tab 3 \tab 3 \tab 55 \tab 4 \tab 5 \tab 7 \tab 4 \tab 13 \tab 11 \tab 10 \tab 12 \tab 2 \tab 6 \tab 29 \tab 2 \tab 14 \tab - \tab 5 \tab 8 \tab 6 \tab 5 \tab 6 \tab 1 \tab 39 \tab 1 \tab 7 \tab 15 \tab 3 \tab 3 \tab 14 \tab 19 \tab 5 \tab 47 \tab - \tab 3 \tab 7 \tab 4 \tab - \tab 50 \tab 3 \tab 5 \tab 5 \tab 3 \tab 27 \tab 7 \tab - \tab 44 \tab 1 \tab 5 \tab 2 \tab 6 \tab 15 \tab 4 \tab 35 \tab 2 \tab 1 \tab 6 \tab 4 \tab 91 \tab 6 \tab - \tab 5 \tab 3 \tab 1 \tab 3 \tab 8 \tab 7 \tab 4 \tab 5 \tab 4 \tab 8 \tab 20 \tab 4 \tab 7 \tab - \tab - \tab 1 \tab 10 \tab 3 \tab 33 \tab 13 \tab 4 \tab 6 \tab 38 \tab 236 \tab 3 \tab 1 \tab 14 \tab - \tab - \tab 3 \tab 17 \tab 29 \tab 70 \tab 1 \tab 32 \tab 43 \tab 58 \tab 10 \tab - \tab 8 \tab 3 \tab 3 \tab 5 \tab 12 \tab 6 \tab 2 \tab 5 \tab - \tab 3 \tab 5 \tab 4 \tab 8 \tab 4 \tab 4 \tab 7 \tab 1 \tab 2 \tab 3 \tab 10 \tab 1 \tab 11 \tab 27 \tab 15 \tab 10 \tab 12 \tab 24 \tab 8 \tab - \tab 6 \tab 5 \tab 40 \tab 25 \tab 5 \tab 13 \tab - \tab - \tab 8 \tab 21 \tab 3 \tab 57 \tab 17 \tab 15 \tab 15 \tab 4 \tab 1 \tab 4 \tab 2 \tab 13 \tab 1 \tab 11 \tab 27 \tab 30 \tab - \tab 12 \tab - \tab 1 \tab 1 \tab 23 \tab 4 \tab - \tab 16 \tab 5 \tab 1 \tab 10 \tab 14 \tab 24 \tab 5 \tab 1 \tab 2 \tab 7 \tab 8 \tab 1 \tab 18 \tab 1 \tab 6 \tab 4 \tab 15 \tab 568 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 7 \tab - \tab - \tab 3 \tab 4 \tab - \tab 1 \tab 3 \tab 18 \tab - \tab 3 \tab - \tab 5 \tab - \tab 5 \tab 10 \tab 5 \tab 2 \tab 5 \tab 4 \tab 1 \tab - \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 7 \tab - \tab - \tab 4 \tab 1 \tab - \tab 13 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 6 \tab 1 \tab 6 \tab - \tab - \tab 5 \tab - \tab - \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1 \tab 1 \tab - \tab 3 \tab 7 \tab 52 \tab 6 \tab 1 \tab - \tab 3 \tab 14 \tab 15 \tab 6 \tab 10 \tab 3 \tab 4 \tab 31 \tab - \tab 6 \tab 1 \tab 4 \tab 3 \tab 4 \tab 2 \tab 4 \tab 2 \tab 28 \tab 4 \tab 7 \tab 12 \tab 3 \tab 1 \tab 23 \tab 29 \tab 3 \tab 25 \tab 1 \tab 4 \tab 3 \tab 2 \tab 7 \tab 50 \tab 2 \tab 3 \tab 1 \tab 4 \tab 15 \tab 3 \tab - \tab 55 \tab 3 \tab 2 \tab 2 \tab 5 \tab 14 \tab 7 \tab 32 \tab 2 \tab 1 \tab 10 \tab 2 \tab 80 \tab 6 \tab - \tab 3 \tab - \tab - \tab 3 \tab 11 \tab 9 \tab 4 \tab 15 \tab 6 \tab 6 \tab 17 \tab 8 \tab 6 \tab - \tab - \tab 1 \tab 22 \tab 2 \tab 38 \tab 10 \tab 6 \tab 8 \tab 29 \tab 228 \tab 3 \tab 2 \tab 22 \tab 2 \tab - \tab 3 \tab 22 \tab 20 \tab 73 \tab 1 \tab 29 \tab 37 \tab 56 \tab 11 \tab 3 \tab 10 \tab 2 \tab 2 \tab - \tab 14 \tab 2 \tab 2 \tab 11 \tab 1 \tab 2 \tab 3 \tab 4 \tab 14 \tab 5 \tab 4 \tab 3 \tab 3 \tab 3 \tab 4 \tab 12 \tab 2 \tab 4 \tab 19 \tab 21 \tab 10 \tab 16 \tab 28 \tab 7 \tab 2 \tab 4 \tab 4 \tab 25 \tab 18 \tab 3 \tab 14 \tab 2 \tab 2 \tab 7 \tab 24 \tab 1 \tab 51 \tab 14 \tab 15 \tab 14 \tab - \tab 3 \tab 1 \tab 1 \tab 11 \tab 3 \tab 9 \tab 22 \tab 19 \tab 1 \tab 6 \tab 1 \tab 6 \tab - \tab 21 \tab 2 \tab - \tab 16 \tab 4 \tab 1 \tab 22 \tab 16 \tab 20 \tab 5 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 23 \tab 1 \tab 6 \tab 5 \tab 17 \tab 569 \tab 5 \tab 1 \tab 4 \tab 1 \tab 4 \tab 3 \tab 1 \tab 3 \tab 1 \tab 9 \tab 2 \tab - \tab 7 \tab 1 \tab 4 \tab 3 \tab 5 \tab 8 \tab 1 \tab - \tab - \tab 5 \tab 1 \tab 4 \tab 15 \tab 6 \tab 2 \tab 5 \tab - \tab 2 \tab - \tab 1 \tab 5 \tab 2 \tab 1 \tab 5 \tab - \tab - \tab 2 \tab 3 \tab 10 \tab - \tab - \tab - \tab 1 \tab 1 \tab 11 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 7 \tab 1 \tab 8 \tab - \tab 1 \tab 6 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 4 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1 \tab 2 \tab 2 \tab 5 \tab 5 \tab 59 \tab 3 \tab 2 \tab - \tab 3 \tab 15 \tab 14 \tab 7 \tab 10 \tab 2 \tab 5 \tab 33 \tab 2 \tab 13 \tab 1 \tab 2 \tab 5 \tab 2 \tab 1 \tab 10 \tab 1 \tab 39 \tab 6 \tab 10 \tab 8 \tab 2 \tab 1 \tab 14 \tab 16 \tab 7 \tab 38 \tab 1 \tab 2 \tab 5 \tab 8 \tab 2 \tab 51 \tab 7 \tab 2 \tab 7 \tab 1 \tab 20 \tab 8 \tab 1 \tab 46 \tab 4 \tab - \tab 1 \tab 6 \tab 20 \tab 10 \tab 41 \tab 3 \tab - \tab 5 \tab 1 \tab 120 \tab 10 \tab 1 \tab 2 \tab - \tab 3 \tab 6 \tab 8 \tab 10 \tab 2 \tab 15 \tab 9 \tab 8 \tab 22 \tab 7 \tab 7 \tab - \tab 4 \tab 1 \tab 13 \tab 1 \tab 24 \tab 15 \tab 4 \tab 4 \tab 25 \tab 246 \tab - \tab 2 \tab 11 \tab - \tab - \tab 1 \tab 20 \tab 14 \tab 64 \tab 1 \tab 18 \tab 30 \tab 50 \tab 14 \tab - \tab 6 \tab 4 \tab 1 \tab 3 \tab 14 \tab 3 \tab 1 \tab 10 \tab - \tab - \tab 7 \tab 3 \tab 11 \tab 4 \tab 3 \tab 4 \tab 2 \tab 5 \tab 3 \tab 9 \tab - \tab 5 \tab 20 \tab 15 \tab 8 \tab 8 \tab 39 \tab 4 \tab 3 \tab 6 \tab 5 \tab 39 \tab 22 \tab 5 \tab 4 \tab 1 \tab 1 \tab 13 \tab 19 \tab 1 \tab 48 \tab 10 \tab 9 \tab 11 \tab - \tab 7 \tab 5 \tab 4 \tab 7 \tab 2 \tab 10 \tab 21 \tab 22 \tab - \tab 5 \tab 1 \tab 5 \tab - \tab 16 \tab 1 \tab - \tab 13 \tab 3 \tab 2 \tab 16 \tab 13 \tab 19 \tab 4 \tab 1 \tab 1 \tab 4 \tab 8 \tab 1 \tab 13 \tab - \tab 4 \tab 4 \tab 13 \tab 576 \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 5 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 4 \tab 11 \tab 1 \tab 6 \tab - \tab 3 \tab 1 \tab 9 \tab 11 \tab 8 \tab 3 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 16 \tab 1 \tab - \tab - \tab 1 \tab - \tab 9 \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 10 \tab - \tab 5 \tab 1 \tab - \tab 10 \tab 1 \tab - \tab 3 \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1 \tab 2 \tab 1 \tab 2 \tab 8 \tab 54 \tab 4 \tab 1 \tab 1 \tab 3 \tab 13 \tab 11 \tab 6 \tab 13 \tab 1 \tab 1 \tab 25 \tab - \tab 6 \tab 1 \tab - \tab 5 \tab 4 \tab 1 \tab 6 \tab - \tab 34 \tab 2 \tab 6 \tab 9 \tab 4 \tab 2 \tab 12 \tab 12 \tab 8 \tab 39 \tab 1 \tab 7 \tab 4 \tab 4 \tab 6 \tab 66 \tab 5 \tab 1 \tab 2 \tab 8 \tab 19 \tab 9 \tab - \tab 55 \tab 3 \tab 7 \tab 1 \tab 3 \tab 12 \tab 6 \tab 41 \tab 2 \tab - \tab 4 \tab 1 \tab 97 \tab 10 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 17 \tab 3 \tab 4 \tab 13 \tab 8 \tab 7 \tab 16 \tab 14 \tab 4 \tab 1 \tab 2 \tab - \tab 13 \tab 3 \tab 17 \tab 10 \tab 8 \tab 6 \tab 27 \tab 203 \tab 2 \tab 2 \tab 19 \tab - \tab - \tab 2 \tab 17 \tab 15 \tab 64 \tab 2 \tab 30 \tab 34 \tab 50 \tab 9 \tab 1 \tab 13 \tab 6 \tab 1 \tab 3 \tab 13 \tab 3 \tab 3 \tab 23 \tab - \tab 2 \tab 5 \tab 4 \tab 14 \tab 4 \tab 3 \tab 8 \tab - \tab 3 \tab 2 \tab 9 \tab 2 \tab 6 \tab 24 \tab 8 \tab 5 \tab 16 \tab 28 \tab 5 \tab 2 \tab 6 \tab 1 \tab 32 \tab 18 \tab 3 \tab 15 \tab - \tab 1 \tab 8 \tab 28 \tab 1 \tab 47 \tab 13 \tab 14 \tab 13 \tab 2 \tab 4 \tab 4 \tab 6 \tab 12 \tab 1 \tab 13 \tab 26 \tab 17 \tab 1 \tab 2 \tab 1 \tab 5 \tab 1 \tab 18 \tab - \tab 1 \tab 13 \tab 3 \tab - \tab 16 \tab 15 \tab 12 \tab 5 \tab - \tab 2 \tab 7 \tab 5 \tab 1 \tab 13 \tab - \tab 4 \tab 5 \tab 13 \tab 589 \tab - \tab 1 \tab 5 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 7 \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab - \tab 4 \tab 13 \tab 3 \tab 2 \tab - \tab 2 \tab 1 \tab 5 \tab 16 \tab 3 \tab 4 \tab 8 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2 \tab 3 \tab 3 \tab 4 \tab - \tab 2 \tab 4 \tab 8 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 9 \tab 2 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab 8 \tab 1 \tab - \tab - \tab 4 \tab 3 \tab - \tab 1 \tab 5 \tab - \tab 9 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1 \tab - \tab - \tab 2 \tab 5 \tab 62 \tab 5 \tab - \tab 1 \tab 3 \tab 14 \tab 15 \tab 7 \tab 14 \tab 3 \tab 4 \tab 30 \tab - \tab 9 \tab 1 \tab 2 \tab 3 \tab 5 \tab 3 \tab 5 \tab - \tab 23 \tab 6 \tab 4 \tab 18 \tab 5 \tab 5 \tab 13 \tab 18 \tab 9 \tab 54 \tab - \tab 8 \tab 4 \tab 4 \tab 3 \tab 50 \tab - \tab 2 \tab 6 \tab 6 \tab 20 \tab 9 \tab 1 \tab 56 \tab 2 \tab 1 \tab 4 \tab 6 \tab 11 \tab 7 \tab 31 \tab 2 \tab 1 \tab 13 \tab 4 \tab 93 \tab 10 \tab - \tab 4 \tab - \tab - \tab 6 \tab 9 \tab 4 \tab 4 \tab 12 \tab 5 \tab 8 \tab 20 \tab 9 \tab 2 \tab 1 \tab - \tab - \tab 15 \tab 2 \tab 28 \tab 8 \tab 5 \tab 10 \tab 44 \tab 174 \tab 1 \tab 1 \tab 8 \tab 1 \tab - \tab 3 \tab 29 \tab 19 \tab 56 \tab - \tab 21 \tab 24 \tab 57 \tab 14 \tab - \tab 10 \tab 7 \tab 6 \tab 3 \tab 23 \tab 1 \tab 2 \tab 21 \tab - \tab 2 \tab 1 \tab 2 \tab 8 \tab 7 \tab 2 \tab 4 \tab - \tab 5 \tab 1 \tab 6 \tab - \tab 6 \tab 18 \tab 15 \tab 3 \tab 16 \tab 31 \tab 4 \tab 3 \tab 2 \tab 6 \tab 27 \tab 15 \tab 3 \tab 12 \tab 1 \tab 2 \tab 9 \tab 23 \tab 2 \tab 48 \tab 16 \tab 9 \tab 15 \tab 1 \tab 1 \tab 3 \tab 1 \tab 14 \tab - \tab 10 \tab 28 \tab 21 \tab 1 \tab 5 \tab 1 \tab 4 \tab 4 \tab 24 \tab - \tab - \tab 24 \tab 5 \tab 2 \tab 23 \tab 14 \tab 12 \tab 7 \tab 1 \tab - \tab 4 \tab 5 \tab 2 \tab 20 \tab - \tab 1 \tab 2 \tab 20 \tab 603 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab 3 \tab 2 \tab 7 \tab 1 \tab 1 \tab 2 \tab 4 \tab 3 \tab - \tab 3 \tab 11 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab 14 \tab 9 \tab - \tab 2 \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 14 \tab - \tab 1 \tab - \tab 1 \tab - \tab 15 \tab 5 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 5 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 6 \tab 1 \tab 5 \tab 3 \tab 7 \tab 2 \tab 2 \tab 5 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 6 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 1 \tab 1 \tab 5 \tab 3 \tab 6 \tab 79 \tab 10 \tab 2 \tab 3 \tab 3 \tab 14 \tab 17 \tab 2 \tab 22 \tab 4 \tab 8 \tab 33 \tab 1 \tab 13 \tab 2 \tab 3 \tab 5 \tab 4 \tab 4 \tab 4 \tab 2 \tab 38 \tab 2 \tab 10 \tab 15 \tab 7 \tab 5 \tab 17 \tab 26 \tab 7 \tab 54 \tab 2 \tab 5 \tab 4 \tab 6 \tab 4 \tab 53 \tab 10 \tab 1 \tab 3 \tab 6 \tab 25 \tab 3 \tab - \tab 53 \tab - \tab 3 \tab 2 \tab 7 \tab 15 \tab 6 \tab 37 \tab 1 \tab - \tab 10 \tab 1 \tab 102 \tab 6 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 12 \tab 5 \tab 4 \tab 11 \tab 5 \tab 5 \tab 24 \tab 8 \tab 3 \tab - \tab 2 \tab - \tab 14 \tab 2 \tab 21 \tab 21 \tab 7 \tab 6 \tab 35 \tab 193 \tab 1 \tab 2 \tab 14 \tab - \tab - \tab - \tab 18 \tab 16 \tab 66 \tab - \tab 31 \tab 48 \tab 40 \tab 8 \tab 3 \tab 7 \tab 7 \tab 2 \tab 4 \tab 13 \tab 1 \tab 1 \tab 31 \tab - \tab - \tab 4 \tab 5 \tab 18 \tab 10 \tab 4 \tab 4 \tab 2 \tab - \tab 3 \tab 7 \tab 1 \tab 6 \tab 18 \tab 21 \tab 11 \tab 14 \tab 26 \tab 9 \tab 3 \tab 4 \tab 5 \tab 32 \tab 24 \tab 5 \tab 10 \tab 2 \tab 1 \tab 11 \tab 19 \tab 1 \tab 55 \tab 10 \tab 16 \tab 17 \tab - \tab 3 \tab 3 \tab 3 \tab 6 \tab 2 \tab 16 \tab 24 \tab 22 \tab 2 \tab 2 \tab 1 \tab 4 \tab - \tab 10 \tab - \tab - \tab 18 \tab 5 \tab - \tab 12 \tab 15 \tab 16 \tab 7 \tab 1 \tab 2 \tab 9 \tab 5 \tab 1 \tab 22 \tab 1 \tab 8 \tab 5 \tab 20 \tab 2107 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 4 \tab 2 \tab 4 \tab 5 \tab 5 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 5 \tab 17 \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab 4 \tab 14 \tab 2 \tab 5 \tab 6 \tab 1 \tab 2 \tab 1 \tab - \tab 5 \tab 2 \tab - \tab 6 \tab 1 \tab - \tab 1 \tab 1 \tab 13 \tab 2 \tab - \tab - \tab - \tab - \tab 23 \tab 3 \tab 3 \tab 4 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 2 \tab 1 \tab 2 \tab 2 \tab 5 \tab 3 \tab 1 \tab 7 \tab - \tab 4 \tab 2 \tab - \tab 3 \tab 2 \tab - \tab 2 \tab - \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 1 \tab - \tab 2 \tab 5 \tab 35 \tab 7 \tab - \tab 3 \tab 1 \tab 12 \tab 11 \tab 8 \tab 9 \tab 2 \tab 4 \tab 24 \tab - \tab 10 \tab 4 \tab 2 \tab 4 \tab - \tab 2 \tab - \tab 2 \tab 22 \tab 3 \tab 8 \tab 8 \tab 4 \tab 1 \tab 5 \tab 6 \tab 2 \tab 32 \tab - \tab 7 \tab 4 \tab 2 \tab 1 \tab 44 \tab 2 \tab 2 \tab 3 \tab 5 \tab 15 \tab 4 \tab - \tab 23 \tab 1 \tab 4 \tab 1 \tab 9 \tab 9 \tab 5 \tab 27 \tab 2 \tab 1 \tab 5 \tab - \tab 64 \tab 8 \tab 1 \tab 3 \tab - \tab 1 \tab 4 \tab 6 \tab 2 \tab 2 \tab 10 \tab 2 \tab 5 \tab 18 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 8 \tab 1 \tab 33 \tab 5 \tab 4 \tab 3 \tab 20 \tab 128 \tab 6 \tab 2 \tab 11 \tab 1 \tab - \tab 3 \tab 18 \tab 11 \tab 26 \tab 1 \tab 28 \tab 26 \tab 32 \tab 13 \tab 1 \tab 13 \tab 1 \tab - \tab 2 \tab 13 \tab 3 \tab 1 \tab 25 \tab - \tab 2 \tab 3 \tab 1 \tab 8 \tab 7 \tab 3 \tab 5 \tab 1 \tab 3 \tab 2 \tab 11 \tab 2 \tab 3 \tab 11 \tab 23 \tab 6 \tab 7 \tab 24 \tab 1 \tab 1 \tab 3 \tab 4 \tab 18 \tab 11 \tab 3 \tab 11 \tab 1 \tab 1 \tab 10 \tab 15 \tab 2 \tab 49 \tab 10 \tab 11 \tab 5 \tab 5 \tab 2 \tab 3 \tab 1 \tab 8 \tab 1 \tab 6 \tab 18 \tab 17 \tab 2 \tab 3 \tab 3 \tab 4 \tab - \tab 8 \tab 2 \tab 1 \tab 17 \tab 2 \tab 2 \tab 9 \tab 7 \tab 18 \tab 3 \tab - \tab 1 \tab 1 \tab 5 \tab 1 \tab 8 \tab 3 \tab 3 \tab 2 \tab 10 \tab 398 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab 4 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 11 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 3 \tab 14 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 4 \tab 3 \tab - \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 9 \tab - \tab - \tab - \tab 1 \tab - \tab 6 \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 6 \tab - \tab - \tab - \tab 2 \tab 1 \tab 6 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab 6 \tab 1 \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 4 \tab - \tab 2 \tab 7 \tab 39 \tab 6 \tab 1 \tab 1 \tab 3 \tab 11 \tab 6 \tab 2 \tab 7 \tab 2 \tab 3 \tab 26 \tab - \tab 7 \tab - \tab 4 \tab 5 \tab 3 \tab 4 \tab 2 \tab 3 \tab 24 \tab 2 \tab 6 \tab 8 \tab 4 \tab 2 \tab 10 \tab 13 \tab 7 \tab 39 \tab 3 \tab 4 \tab 3 \tab 5 \tab 1 \tab 25 \tab 2 \tab - \tab 3 \tab 4 \tab 13 \tab 10 \tab - \tab 31 \tab 2 \tab 3 \tab 2 \tab 8 \tab 15 \tab 3 \tab 31 \tab 3 \tab - \tab 6 \tab - \tab 75 \tab 8 \tab 2 \tab 6 \tab - \tab - \tab 3 \tab 4 \tab 4 \tab 4 \tab 6 \tab 7 \tab 1 \tab 18 \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 16 \tab 2 \tab 17 \tab 13 \tab 4 \tab 4 \tab 21 \tab 115 \tab 1 \tab 4 \tab 14 \tab - \tab 1 \tab - \tab 11 \tab 10 \tab 50 \tab 1 \tab 18 \tab 26 \tab 33 \tab 7 \tab - \tab 6 \tab 7 \tab - \tab - \tab 17 \tab 4 \tab - \tab 18 \tab - \tab 2 \tab 6 \tab 4 \tab 1 \tab 3 \tab 3 \tab 4 \tab - \tab 3 \tab 1 \tab 6 \tab 1 \tab 2 \tab 15 \tab 10 \tab 3 \tab 5 \tab 16 \tab 5 \tab 6 \tab 2 \tab 8 \tab 24 \tab 7 \tab 3 \tab 11 \tab - \tab 1 \tab 4 \tab 13 \tab - \tab 49 \tab 11 \tab 11 \tab 14 \tab - \tab 4 \tab 2 \tab 4 \tab 8 \tab - \tab 4 \tab 14 \tab 20 \tab - \tab 4 \tab 1 \tab 3 \tab 1 \tab 11 \tab - \tab - \tab 11 \tab 2 \tab 1 \tab 19 \tab 11 \tab 14 \tab 2 \tab 3 \tab 1 \tab 2 \tab 7 \tab - \tab 11 \tab 1 \tab 3 \tab 7 \tab 20 \tab 402 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 3 \tab 2 \tab - \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 14 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 5 \tab 14 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 7 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 8 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab 4 \tab - \tab 5 \tab - \tab 8 \tab - \tab - \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab 1 \tab 2 \tab 2 \tab 1 \tab 7 \tab 51 \tab 2 \tab - \tab 2 \tab 2 \tab 10 \tab 7 \tab 3 \tab 6 \tab 1 \tab 1 \tab 31 \tab 1 \tab 7 \tab 2 \tab 2 \tab 3 \tab 2 \tab 2 \tab 4 \tab 2 \tab 29 \tab 6 \tab 6 \tab 11 \tab 2 \tab 1 \tab 10 \tab 12 \tab 3 \tab 55 \tab - \tab 4 \tab 5 \tab - \tab 1 \tab 55 \tab 4 \tab - \tab 4 \tab 2 \tab 17 \tab 3 \tab - \tab 30 \tab 1 \tab 1 \tab - \tab 1 \tab 8 \tab 5 \tab 22 \tab - \tab - \tab 4 \tab - \tab 57 \tab 8 \tab - \tab 4 \tab - \tab 1 \tab 2 \tab 5 \tab 5 \tab - \tab 10 \tab 2 \tab 3 \tab 12 \tab 9 \tab 1 \tab 1 \tab 1 \tab 2 \tab 7 \tab - \tab 23 \tab 14 \tab - \tab 9 \tab 19 \tab 114 \tab 1 \tab - \tab 6 \tab - \tab - \tab - \tab 11 \tab 7 \tab 39 \tab - \tab 16 \tab 20 \tab 22 \tab 3 \tab - \tab 4 \tab 2 \tab 3 \tab - \tab 9 \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 4 \tab 5 \tab 2 \tab 3 \tab 5 \tab - \tab 4 \tab 3 \tab 2 \tab - \tab - \tab 6 \tab 15 \tab 4 \tab 11 \tab 26 \tab 4 \tab 1 \tab 6 \tab 1 \tab 24 \tab 19 \tab 2 \tab 52 \tab 1 \tab - \tab 12 \tab 16 \tab 1 \tab 53 \tab 6 \tab 9 \tab 7 \tab 4 \tab 1 \tab 2 \tab 1 \tab 7 \tab - \tab 7 \tab 10 \tab 15 \tab 2 \tab 4 \tab 1 \tab 4 \tab - \tab 15 \tab 3 \tab 2 \tab 13 \tab 3 \tab 1 \tab 21 \tab 10 \tab 11 \tab 2 \tab 1 \tab 2 \tab - \tab 4 \tab 2 \tab 8 \tab 2 \tab 2 \tab 3 \tab 22 \tab 385 \tab 1 \tab 2 \tab 5 \tab 1 \tab 2 \tab - \tab - \tab 4 \tab - \tab 2 \tab 2 \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab 8 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 8 \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 23 \tab - \tab - \tab 2 \tab 1 \tab - \tab 7 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 5 \tab 1 \tab 6 \tab - \tab 8 \tab 1 \tab 1 \tab 4 \tab 4 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1 \tab 1 \tab - \tab 2 \tab 5 \tab 58 \tab 3 \tab 3 \tab - \tab 5 \tab 14 \tab 13 \tab 5 \tab 6 \tab 3 \tab 6 \tab 30 \tab 2 \tab 9 \tab - \tab 1 \tab 2 \tab 2 \tab 3 \tab 3 \tab 1 \tab 33 \tab 2 \tab 5 \tab 16 \tab 3 \tab - \tab 12 \tab 17 \tab 3 \tab 51 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 54 \tab 1 \tab - \tab 4 \tab 8 \tab 21 \tab 10 \tab - \tab 43 \tab - \tab 4 \tab 2 \tab 2 \tab 14 \tab 3 \tab 34 \tab 1 \tab - \tab 7 \tab 1 \tab 80 \tab 6 \tab 2 \tab 5 \tab - \tab - \tab 2 \tab 9 \tab 3 \tab 5 \tab 20 \tab 3 \tab 6 \tab 28 \tab 6 \tab 4 \tab - \tab 1 \tab - \tab 13 \tab 3 \tab 20 \tab 12 \tab 3 \tab 6 \tab 36 \tab 132 \tab 1 \tab 2 \tab 19 \tab - \tab 1 \tab - \tab 17 \tab 6 \tab 57 \tab - \tab 16 \tab 27 \tab 35 \tab 12 \tab - \tab 10 \tab 2 \tab 5 \tab - \tab 8 \tab 3 \tab 3 \tab 42 \tab - \tab 1 \tab 4 \tab 2 \tab 10 \tab 4 \tab 5 \tab 8 \tab - \tab 11 \tab 3 \tab 3 \tab 1 \tab 7 \tab 17 \tab 12 \tab 1 \tab 10 \tab 32 \tab 2 \tab 3 \tab 2 \tab 4 \tab 46 \tab 19 \tab 3 \tab 7 \tab - \tab - \tab 7 \tab 28 \tab 1 \tab 30 \tab 23 \tab 17 \tab 14 \tab 3 \tab 2 \tab 1 \tab 6 \tab 5 \tab 1 \tab 5 \tab 23 \tab 25 \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab 14 \tab 2 \tab - \tab 19 \tab 4 \tab 2 \tab 19 \tab 14 \tab 13 \tab 1 \tab 1 \tab 2 \tab 5 \tab 3 \tab - \tab 19 \tab - \tab 1 \tab 4 \tab 20 \tab 524 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 7 \tab 14 \tab 1 \tab - \tab 1 \tab 8 \tab - \tab 7 \tab 12 \tab 4 \tab 2 \tab 6 \tab - \tab 3 \tab 1 \tab 2 \tab 6 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 4 \tab 2 \tab - \tab - \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 3 \tab - \tab 9 \tab - \tab 10 \tab 1 \tab - \tab 7 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 6 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2538 \cr
   2016 \tab - \tab 1 \tab - \tab 3 \tab 10 \tab 79 \tab 1 \tab 1 \tab 2 \tab 6 \tab 13 \tab 8 \tab 8 \tab 8 \tab 2 \tab 7 \tab 53 \tab 4 \tab 9 \tab - \tab 5 \tab 8 \tab 4 \tab 5 \tab 3 \tab - \tab 36 \tab 2 \tab 6 \tab 20 \tab 4 \tab 5 \tab 7 \tab 13 \tab 8 \tab 70 \tab 1 \tab 4 \tab 2 \tab 6 \tab 3 \tab 59 \tab 4 \tab 1 \tab 3 \tab 10 \tab 27 \tab 10 \tab - \tab 40 \tab 3 \tab - \tab 2 \tab 2 \tab 13 \tab 9 \tab 41 \tab 1 \tab - \tab 9 \tab 3 \tab 93 \tab 9 \tab 1 \tab 6 \tab - \tab - \tab 2 \tab 9 \tab 3 \tab 7 \tab 13 \tab 9 \tab 5 \tab 24 \tab 9 \tab 3 \tab - \tab 5 \tab - \tab 15 \tab 4 \tab 27 \tab 7 \tab 5 \tab 7 \tab 33 \tab 119 \tab 4 \tab 5 \tab 16 \tab 1 \tab - \tab 2 \tab 13 \tab 7 \tab 51 \tab 1 \tab 25 \tab 17 \tab 36 \tab 15 \tab - \tab 6 \tab 7 \tab 3 \tab 4 \tab 13 \tab 6 \tab 3 \tab 24 \tab 1 \tab 2 \tab 5 \tab 4 \tab 10 \tab 4 \tab 8 \tab 3 \tab 2 \tab 8 \tab 9 \tab 6 \tab 1 \tab 13 \tab 18 \tab 11 \tab 5 \tab 8 \tab 27 \tab 5 \tab 2 \tab 9 \tab 6 \tab 20 \tab 14 \tab 6 \tab 3 \tab - \tab 1 \tab 7 \tab 36 \tab 1 \tab 57 \tab 12 \tab 22 \tab 18 \tab 2 \tab 5 \tab 1 \tab 4 \tab 18 \tab 2 \tab 6 \tab 25 \tab 20 \tab 1 \tab 4 \tab 2 \tab 8 \tab 1 \tab 21 \tab 1 \tab - \tab 32 \tab 2 \tab 1 \tab 16 \tab 15 \tab 11 \tab 2 \tab 5 \tab 4 \tab 6 \tab 5 \tab - \tab 16 \tab 1 \tab 5 \tab 3 \tab 17 \tab 625 \tab - \tab 3 \tab 4 \tab 2 \tab 1 \tab 4 \tab 3 \tab 4 \tab 3 \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 7 \tab 11 \tab 3 \tab 1 \tab - \tab 5 \tab - \tab 7 \tab 12 \tab 9 \tab 8 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 9 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 11 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 4 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 5 \tab 1 \tab 11 \tab - \tab 9 \tab - \tab - \tab 9 \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab 4 \tab 6 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 5 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2867 \cr
   2018 \tab - \tab 2 \tab 2 \tab 2 \tab 10 \tab 52 \tab 2 \tab 1 \tab 2 \tab 4 \tab 14 \tab 6 \tab 6 \tab 8 \tab 2 \tab 5 \tab 33 \tab 4 \tab 5 \tab 2 \tab 5 \tab 3 \tab 4 \tab 9 \tab 6 \tab 1 \tab 30 \tab 5 \tab 4 \tab 8 \tab 5 \tab 1 \tab 7 \tab 10 \tab 6 \tab 37 \tab - \tab 2 \tab 4 \tab 4 \tab 1 \tab 37 \tab 3 \tab 1 \tab 3 \tab 6 \tab 23 \tab 8 \tab - \tab 34 \tab 1 \tab 2 \tab 3 \tab 2 \tab 10 \tab 4 \tab 32 \tab 6 \tab 1 \tab 9 \tab 3 \tab 85 \tab 6 \tab 2 \tab - \tab 1 \tab - \tab 6 \tab 3 \tab 6 \tab 4 \tab 9 \tab 9 \tab 4 \tab 11 \tab 16 \tab 1 \tab - \tab 1 \tab - \tab 9 \tab 2 \tab 11 \tab 11 \tab 38 \tab 3 \tab 30 \tab 100 \tab 2 \tab 3 \tab 4 \tab 1 \tab 1 \tab 1 \tab 9 \tab 10 \tab 31 \tab 1 \tab 11 \tab 48 \tab 26 \tab 10 \tab - \tab 9 \tab 10 \tab 3 \tab 2 \tab 14 \tab 4 \tab 1 \tab 25 \tab 1 \tab 2 \tab 3 \tab 5 \tab 19 \tab 2 \tab 2 \tab 7 \tab - \tab 4 \tab 7 \tab 10 \tab - \tab 15 \tab 7 \tab 17 \tab 3 \tab 12 \tab 28 \tab 7 \tab 2 \tab 2 \tab 7 \tab 21 \tab 11 \tab 2 \tab 2 \tab - \tab 3 \tab 15 \tab 26 \tab 3 \tab 26 \tab 7 \tab 16 \tab 10 \tab - \tab 2 \tab 2 \tab 5 \tab 10 \tab 1 \tab 8 \tab 20 \tab 28 \tab 3 \tab 3 \tab - \tab 4 \tab - \tab 12 \tab 3 \tab 1 \tab 10 \tab 3 \tab 2 \tab 13 \tab 10 \tab 5 \tab - \tab 2 \tab 2 \tab 2 \tab 4 \tab - \tab 9 \tab 2 \tab 3 \tab 6 \tab 19 \tab 498 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 5 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 5 \tab 11 \tab 2 \tab - \tab 1 \tab 4 \tab - \tab 10 \tab 7 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 12 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab 3 \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 4 \tab 1 \tab 5 \tab - \tab - \tab - \tab 8 \tab 4 \tab 2 \tab 2 \tab 8 \tab 1 \tab 5 \tab 1 \tab 1 \tab 8 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab 4 \tab 6 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2348 \cr
   2021 \tab - \tab 3 \tab 3 \tab 3 \tab 15 \tab 50 \tab 1 \tab 3 \tab 3 \tab 5 \tab 28 \tab 20 \tab 6 \tab 10 \tab 2 \tab 12 \tab 47 \tab 3 \tab 16 \tab 3 \tab 2 \tab 12 \tab 9 \tab 6 \tab 3 \tab 1 \tab 51 \tab 9 \tab 8 \tab 18 \tab 3 \tab 3 \tab 13 \tab 38 \tab 11 \tab 74 \tab - \tab 5 \tab 8 \tab 2 \tab 3 \tab 79 \tab 9 \tab 1 \tab 9 \tab 6 \tab 32 \tab 12 \tab - \tab 80 \tab 1 \tab 1 \tab 4 \tab 6 \tab 19 \tab 15 \tab 50 \tab - \tab - \tab 20 \tab 2 \tab 151 \tab 12 \tab - \tab 2 \tab - \tab - \tab 2 \tab 12 \tab 11 \tab 6 \tab 14 \tab 5 \tab 15 \tab 42 \tab 20 \tab 11 \tab 1 \tab 1 \tab 1 \tab 22 \tab 6 \tab 55 \tab 25 \tab - \tab 11 \tab 55 \tab 122 \tab 2 \tab 4 \tab 22 \tab 1 \tab 3 \tab 1 \tab 15 \tab 23 \tab 102 \tab 3 \tab 52 \tab 31 \tab 52 \tab 9 \tab 1 \tab 10 \tab 3 \tab 4 \tab 4 \tab 18 \tab 1 \tab 7 \tab 82 \tab 1 \tab 3 \tab 6 \tab 8 \tab 12 \tab 6 \tab 4 \tab 11 \tab 1 \tab 7 \tab 4 \tab 12 \tab 2 \tab 11 \tab 35 \tab 25 \tab 5 \tab 21 \tab 42 \tab 9 \tab 3 \tab 3 \tab 12 \tab 45 \tab 28 \tab 4 \tab - \tab 4 \tab 1 \tab 12 \tab 47 \tab 1 \tab 131 \tab 15 \tab 24 \tab 27 \tab 2 \tab 8 \tab 3 \tab 8 \tab 16 \tab - \tab 15 \tab 56 \tab 45 \tab 2 \tab 3 \tab 1 \tab 7 \tab - \tab 25 \tab 3 \tab 1 \tab 38 \tab 8 \tab 5 \tab 31 \tab 23 \tab 17 \tab 6 \tab 4 \tab 3 \tab 10 \tab 15 \tab - \tab 33 \tab 2 \tab 6 \tab 12 \tab 31 \tab 667 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 11 \tab 3 \tab 10 \tab 1 \tab 4 \tab 4 \tab 1 \tab - \tab 1 \tab 6 \tab 23 \tab 4 \tab 1 \tab - \tab 6 \tab 2 \tab 6 \tab 15 \tab 7 \tab 3 \tab 12 \tab 2 \tab 6 \tab - \tab - \tab 10 \tab 10 \tab 1 \tab 6 \tab - \tab - \tab 3 \tab 2 \tab - \tab 5 \tab - \tab 3 \tab 1 \tab - \tab 21 \tab 5 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 8 \tab - \tab 1 \tab - \tab 9 \tab 9 \tab 2 \tab 1 \tab 11 \tab 4 \tab 28 \tab 2 \tab 2 \tab 9 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab 7 \tab 3 \tab 3 \tab 2 \tab 3 \tab 5 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 6 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 9 \tab 6 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 5 \tab 9 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab 3 \tab 1 \tab 9 \tab 61 \tab 1 \tab 2 \tab - \tab 5 \tab 30 \tab 10 \tab 8 \tab 6 \tab 3 \tab 12 \tab 37 \tab 4 \tab 5 \tab 2 \tab 4 \tab 5 \tab 6 \tab 2 \tab 5 \tab 1 \tab 61 \tab 6 \tab 2 \tab 15 \tab 8 \tab 5 \tab 13 \tab 32 \tab 7 \tab 95 \tab 2 \tab 2 \tab 3 \tab 4 \tab - \tab 66 \tab 3 \tab 1 \tab 7 \tab 9 \tab 18 \tab 7 \tab 1 \tab 65 \tab 1 \tab - \tab 4 \tab 3 \tab 11 \tab 9 \tab 51 \tab 2 \tab 1 \tab 16 \tab 2 \tab 138 \tab 10 \tab - \tab 4 \tab - \tab 2 \tab 4 \tab 10 \tab 7 \tab 5 \tab 14 \tab 5 \tab 9 \tab 17 \tab 14 \tab 6 \tab 1 \tab 2 \tab 2 \tab 12 \tab 2 \tab 31 \tab 16 \tab - \tab 12 \tab 55 \tab 110 \tab - \tab 2 \tab 16 \tab 1 \tab 1 \tab 5 \tab 15 \tab 10 \tab 64 \tab - \tab 39 \tab 30 \tab 35 \tab 7 \tab 1 \tab 9 \tab 5 \tab 3 \tab 5 \tab 8 \tab 7 \tab 8 \tab 78 \tab 1 \tab 4 \tab 7 \tab 2 \tab 4 \tab 7 \tab 5 \tab 8 \tab - \tab 9 \tab 5 \tab 9 \tab 3 \tab 6 \tab 28 \tab 17 \tab 8 \tab 14 \tab 48 \tab 11 \tab 4 \tab - \tab 7 \tab 41 \tab 19 \tab 9 \tab - \tab 1 \tab 5 \tab 18 \tab 49 \tab 3 \tab 91 \tab 9 \tab 15 \tab 25 \tab 3 \tab 3 \tab 1 \tab 6 \tab 7 \tab 3 \tab 12 \tab 32 \tab 31 \tab 1 \tab 4 \tab 2 \tab 8 \tab - \tab 18 \tab - \tab 1 \tab 30 \tab 5 \tab 4 \tab 22 \tab 24 \tab 17 \tab 7 \tab 1 \tab 5 \tab 6 \tab 11 \tab - \tab 26 \tab 2 \tab 8 \tab 4 \tab 33 \tab 752 \tab 1 \tab 5 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab 5 \tab 13 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 6 \tab 21 \tab 1 \tab 2 \tab - \tab 9 \tab 1 \tab 4 \tab 9 \tab 10 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab - \tab - \tab 3 \tab 1 \tab 25 \tab 4 \tab 1 \tab 1 \tab 2 \tab 5 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab 4 \tab 2 \tab 1 \tab - \tab 8 \tab 7 \tab 4 \tab 2 \tab 9 \tab 2 \tab 11 \tab 1 \tab - \tab 8 \tab 3 \tab - \tab 2 \tab 2 \tab 2 \tab 9 \tab 4 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 6 \tab - \tab 2 \tab - \tab - \tab 5 \tab - \tab 4 \tab 3 \tab 1 \tab - \tab 3 \tab 14 \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 5 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab 4 \tab 8 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3544 \cr
   Total \tab 29 \tab 34 \tab 33 \tab 78 \tab 156 \tab 1698 \tab 127 \tab 55 \tab 63 \tab 87 \tab 362 \tab 314 \tab 171 \tab 292 \tab 50 \tab 125 \tab 777 \tab 44 \tab 224 \tab 37 \tab 82 \tab 190 \tab 84 \tab 102 \tab 89 \tab 72 \tab 955 \tab 102 \tab 183 \tab 379 \tab 91 \tab 57 \tab 378 \tab 429 \tab 173 \tab 1074 \tab 28 \tab 165 \tab 124 \tab 141 \tab 82 \tab 1195 \tab 117 \tab 59 \tab 121 \tab 127 \tab 532 \tab 176 \tab 7 \tab 1456 \tab 46 \tab 83 \tab 54 \tab 103 \tab 348 \tab 179 \tab 827 \tab 81 \tab 18 \tab 214 \tab 66 \tab 2324 \tab 195 \tab 26 \tab 77 \tab 16 \tab 17 \tab 95 \tab 240 \tab 142 \tab 109 \tab 304 \tab 175 \tab 172 \tab 530 \tab 209 \tab 193 \tab 15 \tab 39 \tab 14 \tab 354 \tab 50 \tab 742 \tab 335 \tab 125 \tab 184 \tab 792 \tab 6643 \tab 84 \tab 65 \tab 331 \tab 15 \tab 11 \tab 73 \tab 731 \tab 331 \tab 1578 \tab 22 \tab 584 \tab 816 \tab 1197 \tab 311 \tab 19 \tab 238 \tab 119 \tab 67 \tab 44 \tab 437 \tab 86 \tab 63 \tab 554 \tab 9 \tab 41 \tab 83 \tab 93 \tab 275 \tab 131 \tab 107 \tab 118 \tab 22 \tab 100 \tab 74 \tab 223 \tab 26 \tab 153 \tab 470 \tab 404 \tab 161 \tab 316 \tab 611 \tab 182 \tab 57 \tab 159 \tab 120 \tab 814 \tab 462 \tab 105 \tab 259 \tab 19 \tab 23 \tab 238 \tab 591 \tab 30 \tab 1230 \tab 315 \tab 322 \tab 321 \tab 44 \tab 79 \tab 63 \tab 76 \tab 241 \tab 35 \tab 219 \tab 539 \tab 526 \tab 26 \tab 85 \tab 28 \tab 115 \tab 11 \tab 341 \tab 38 \tab 23 \tab 404 \tab 88 \tab 42 \tab 366 \tab 315 \tab 340 \tab 105 \tab 29 \tab 34 \tab 104 \tab 138 \tab 29 \tab 385 \tab 24 \tab 117 \tab 118 \tab 419 \tab 15129 \tab 43 \tab 51 \tab 49 \tab 21 \tab 36 \tab 43 \tab 20 \tab 69 \tab 76 \tab 124 \tab 32 \tab 23 \tab 64 \tab 41 \tab 33 \tab 14 \tab 87 \tab 296 \tab 40 \tab 39 \tab 9 \tab 80 \tab 17 \tab 167 \tab 285 \tab 125 \tab 54 \tab 91 \tab 22 \tab 44 \tab 9 \tab 25 \tab 96 \tab 57 \tab 41 \tab 47 \tab 28 \tab 2 \tab 33 \tab 37 \tab 225 \tab 32 \tab 12 \tab 35 \tab 29 \tab 8 \tab 236 \tab 49 \tab 38 \tab 24 \tab 13 \tab 45 \tab 28 \tab 6 \tab 25 \tab 13 \tab 16 \tab 22 \tab 19 \tab 40 \tab 35 \tab 34 \tab 20 \tab 32 \tab 77 \tab 16 \tab 4 \tab 8 \tab 65 \tab 58 \tab 59 \tab 17 \tab 123 \tab 19 \tab 161 \tab 19 \tab 16 \tab 127 \tab 38 \tab 11 \tab 45 \tab 13 \tab 25 \tab 62 \tab 25 \tab 17 \tab 22 \tab 31 \tab 62 \tab 7 \tab 5 \tab 6 \tab 8 \tab 8 \tab 11 \tab 9 \tab 8 \tab 51 \tab 13 \tab 6 \tab 14 \tab 4 \tab 13 \tab 3 \tab 27 \tab 8 \tab 11 \tab 5 \tab 39 \tab 10 \tab 7 \tab 8 \tab 23 \tab 11 \tab 32 \tab 1 \tab 20 \tab 12 \tab 14 \tab 12 \tab 31 \tab 31 \tab 14 \tab 16 \tab 5 \tab 16 \tab 19 \tab 3 \tab 20 \tab 15 \tab 5 \tab 4 \tab 21 \tab 48 \tab 34 \tab 5 \tab 10 \tab 5 \tab 9 \tab 6 \tab 5 \tab 1 \tab 1 \tab 4 \tab 24 \tab 25 \tab 7 \tab 3 \tab 23 \tab 3 \tab 3 \tab 5 \tab 14 \tab 7 \tab 1 \tab 3 \tab 7 \tab 13 \tab 13 \tab 9 \tab 2 \tab 7 \tab 3 \tab 3 \tab 5 \tab 17 \tab 31 \tab 1 \tab 1 \tab 1 \tab 13 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} i don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
