% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_597.R
\name{pasei10educ}
\alias{pasei10educ}
\title{Percentage of some college educ in paocc10 based on acs 2010}
\source{
General Social Survey https://gss.norc.org
}
\description{
pasei10educ
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 11.9 \tab 12.2 \tab 12.4 \tab 12.6 \tab 13.1 \tab 13.3 \tab 13.5 \tab 14.1 \tab 14.2 \tab 15.5 \tab 16 \tab 16.5 \tab 16.7 \tab 16.8 \tab 16.9 \tab 17 \tab 17.2 \tab 17.3 \tab 18 \tab 18.7 \tab 19.2 \tab 19.3 \tab 19.6 \tab 20.4 \tab 20.5 \tab 20.6 \tab 21 \tab 21.1 \tab 21.6 \tab 21.7 \tab 21.9 \tab 22 \tab 22.3 \tab 22.7 \tab 22.9 \tab 23 \tab 23.9 \tab 24 \tab 24.1 \tab 24.3 \tab 24.4 \tab 25.1 \tab 25.4 \tab 25.6 \tab 25.7 \tab 25.8 \tab 26.2 \tab 27 \tab 27.5 \tab 27.8 \tab 28 \tab 28.2 \tab 28.3 \tab 28.9 \tab 29.5 \tab 30.6 \tab 30.7 \tab 31 \tab 31.2 \tab 31.9 \tab 32.4 \tab 33.7 \tab 33.9 \tab 34.2 \tab 34.8 \tab 34.9 \tab 35.1 \tab 35.3 \tab 35.5 \tab 36 \tab 36.5 \tab 37 \tab 37.4 \tab 37.8 \tab 38.4 \tab 38.5 \tab 39.2 \tab 39.6 \tab 40.2 \tab 40.4 \tab 40.8 \tab 41.6 \tab 42 \tab 42.3 \tab 42.8 \tab 43.9 \tab 44.2 \tab 45.2 \tab 45.4 \tab 45.5 \tab 46.1 \tab 46.4 \tab 46.9 \tab 47.3 \tab 47.8 \tab 47.9 \tab 48.1 \tab 48.2 \tab 48.5 \tab 49 \tab 50 \tab 50.1 \tab 50.7 \tab 52.1 \tab 53 \tab 53.9 \tab 54.2 \tab 55.4 \tab 55.8 \tab 55.9 \tab 56.1 \tab 56.7 \tab 57.7 \tab 58.5 \tab 59.2 \tab 59.3 \tab 60.9 \tab 62.4 \tab 63.1 \tab 63.9 \tab 64.6 \tab 64.7 \tab 65.3 \tab 65.6 \tab 66.9 \tab 67.4 \tab 68.4 \tab 68.6 \tab 69.9 \tab 70.1 \tab 70.4 \tab 72.2 \tab 73 \tab 74.4 \tab 75.5 \tab 75.9 \tab 76.5 \tab 76.9 \tab 77.2 \tab 77.5 \tab 78 \tab 78.1 \tab 78.8 \tab 79.5 \tab 81 \tab 82.3 \tab 82.4 \tab 82.5 \tab 82.8 \tab 83.1 \tab 85.6 \tab 85.7 \tab 86.3 \tab 87 \tab 88.3 \tab 89.4 \tab 90.5 \tab 91.3 \tab 92 \tab 92.9 \tab 93.1 \tab 93.7 \tab 94.4 \tab 95.1 \tab 95.6 \tab 95.7 \tab 96.3 \tab 96.8 \tab 97.3 \tab 98.5 \tab 99 \tab 99.1 \tab 99.4 \tab 99.5 \tab 99.7 \tab iap \tab 12.3 \tab 17.5 \tab 17.7 \tab 17.9 \tab 19.1 \tab 19.7 \tab 19.8 \tab 19.9 \tab 20.8 \tab 21.5 \tab 22.5 \tab 23.6 \tab 26 \tab 26.6 \tab 27.1 \tab 29.1 \tab 29.7 \tab 30.4 \tab 32.1 \tab 32.5 \tab 33.5 \tab 35 \tab 35.2 \tab 37.7 \tab 38.3 \tab 40.3 \tab 47.4 \tab 47.5 \tab 48.8 \tab 50.6 \tab 51.1 \tab 51.7 \tab 53.3 \tab 54.7 \tab 54.8 \tab 55.2 \tab 57.6 \tab 63.3 \tab 63.5 \tab 65.1 \tab 66.8 \tab 67.6 \tab 71.7 \tab 75.6 \tab 78.2 \tab 79.4 \tab 79.6 \tab 81.1 \tab 81.4 \tab 83.3 \tab 84.6 \tab 85.3 \tab 86.8 \tab 87.3 \tab 89.8 \tab 90.9 \tab 95.3 \tab 95.5 \tab 98.9 \tab 99.2 \tab no answer \tab 10.9 \tab 18.6 \tab 19 \tab 22.4 \tab 24.7 \tab 32.8 \tab 33.8 \tab 34.1 \tab 35.8 \tab 36.2 \tab 39.9 \tab 44 \tab 45.7 \tab 48.9 \tab 49.3 \tab 49.7 \tab 54.3 \tab 58.2 \tab 61 \tab 62.2 \tab 64.1 \tab 73.9 \tab 75.1 \tab 78.4 \tab 79.2 \tab 83 \tab 90.6 \tab 91.4 \tab 91.7 \tab 91.9 \tab 95 \tab 97.8 \tab 98.7 \tab 34.3 \tab 42.6 \tab 69.1 \tab 69.7 \tab 69.8 \tab 72.5 \tab 85.9 \tab 92.1 \tab 94 \tab 96.5 \tab 13.4 \tab 52.2 \tab 55.6 \tab 58.9 \tab 68.2 \tab 84.4 \tab 93.3 \tab 94.1 \tab 14.3 \tab 28.7 \tab 61.3 \tab 78.3 \tab 41.8 \tab 46.6 \tab 57.1 \tab 63.7 \tab 66 \tab 69.6 \tab 73.3 \tab 79.9 \tab 80.1 \tab 99.6 \tab 17.8 \tab 54.6 \tab 66.7 \tab 74.2 \tab 83.7 \tab 92.4 \tab 26.7 \tab 52.7 \tab 68.8 \tab 73.5 \tab 80.7 \tab 86.7 \tab 23.2 \tab 29.4 \tab 42.1 \tab 66.3 \tab 74.6 \tab 84 \tab 90.2 \tab 94.8 \tab 61.1 \tab 80 \tab 98.3 \tab 55.7 \tab 81.3 \tab 81.9 \tab 9.8 \tab 43.7 \tab 48.3 \tab 62.7 \tab 65.5 \tab 97 \tab don't know \tab 44.4 \tab 58.6 \tab 79.8 \tab 35.9 \tab 38.6 \tab 55.5 \tab 62.3 \tab 91.1 \tab 83.5 \tab 88.2 \tab 71.6 \tab 90.8 \tab 92.6 \tab 91.2 \tab 96 \tab 96.4 \tab 45.1 \tab 46.3 \tab 67.3 \tab 84.8 \tab 56.6 \tab 60.7 \tab 84.5 \tab 84.7 \tab 62.9 \tab 55.1 \tab 62.8 \tab 97.5 \tab 34.7 \tab 46.7 \tab 58.7 \tab 58.8 \tab 70.9 \tab 70.3 \tab 31.1 \tab 87.5 \tab 94.3 \tab not available in this release \tab Total \cr
   1972 \tab 3 \tab 2 \tab 2 \tab 3 \tab 1 \tab 13 \tab 3 \tab 54 \tab 4 \tab 1 \tab 11 \tab 11 \tab 1 \tab 1 \tab 7 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 10 \tab 3 \tab 2 \tab 14 \tab 2 \tab 4 \tab 6 \tab 1 \tab 11 \tab 7 \tab 1 \tab 3 \tab 23 \tab 3 \tab 30 \tab 2 \tab 3 \tab 2 \tab 19 \tab 7 \tab 1 \tab 7 \tab 38 \tab 6 \tab 35 \tab 1 \tab 1 \tab 2 \tab 1 \tab 15 \tab 3 \tab 2 \tab 4 \tab 7 \tab 18 \tab 3 \tab 2 \tab 4 \tab 12 \tab 5 \tab 7 \tab 4 \tab 14 \tab 3 \tab 2 \tab 1 \tab 11 \tab 1 \tab 1 \tab 3 \tab 3 \tab 8 \tab 4 \tab 7 \tab 1 \tab 5 \tab 1 \tab 6 \tab 1 \tab 3 \tab 1 \tab 5 \tab 25 \tab 12 \tab 1 \tab 6 \tab 1 \tab 300 \tab 3 \tab 11 \tab 7 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 5 \tab 5 \tab 5 \tab 1 \tab 2 \tab 4 \tab 1 \tab 6 \tab 1 \tab 3 \tab 10 \tab 15 \tab 1 \tab 9 \tab 7 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2 \tab 3 \tab 4 \tab 1 \tab 1 \tab 3 \tab 3 \tab 4 \tab 7 \tab 1 \tab 7 \tab 12 \tab 1 \tab 1 \tab 9 \tab 4 \tab 18 \tab 11 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 7 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 11 \tab 2 \tab 1 \tab 3 \tab 1 \tab 7 \tab 9 \tab 6 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 3 \tab 8 \tab 1 \tab 12 \tab 406 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 5 \tab - \tab - \tab 1 \tab - \tab 7 \tab 1 \tab 81 \tab 4 \tab - \tab 11 \tab 4 \tab - \tab 1 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 10 \tab 3 \tab 2 \tab 18 \tab - \tab 1 \tab 6 \tab 7 \tab 12 \tab 8 \tab 1 \tab 5 \tab 14 \tab 2 \tab 34 \tab - \tab 4 \tab 3 \tab 22 \tab 26 \tab 2 \tab 8 \tab 38 \tab - \tab 45 \tab 1 \tab 1 \tab - \tab 1 \tab 8 \tab 3 \tab - \tab 4 \tab 1 \tab 18 \tab 3 \tab 9 \tab 2 \tab 6 \tab 1 \tab 1 \tab 2 \tab 20 \tab 1 \tab 4 \tab 1 \tab 23 \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 6 \tab 1 \tab 8 \tab 1 \tab 7 \tab - \tab 1 \tab 1 \tab 4 \tab 33 \tab 8 \tab 1 \tab 11 \tab - \tab 233 \tab 3 \tab 9 \tab 4 \tab 2 \tab 3 \tab 6 \tab - \tab 2 \tab - \tab 1 \tab 5 \tab 3 \tab 1 \tab 5 \tab 2 \tab - \tab 2 \tab 5 \tab 6 \tab 35 \tab - \tab 16 \tab 7 \tab 1 \tab 2 \tab 7 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 5 \tab - \tab 6 \tab 6 \tab 1 \tab - \tab 2 \tab 2 \tab 7 \tab 3 \tab - \tab 4 \tab 21 \tab - \tab - \tab 8 \tab 5 \tab 17 \tab 8 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 3 \tab 4 \tab - \tab - \tab 7 \tab - \tab - \tab 2 \tab 12 \tab 4 \tab 1 \tab 5 \tab 1 \tab 6 \tab 11 \tab 4 \tab 5 \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 4 \tab 10 \tab - \tab 6 \tab 254 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 3 \tab - \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 63 \tab 2 \tab 2 \tab 9 \tab 6 \tab 2 \tab - \tab 5 \tab - \tab 1 \tab 6 \tab 2 \tab 1 \tab 9 \tab 6 \tab - \tab 13 \tab - \tab 2 \tab 6 \tab 11 \tab 8 \tab 10 \tab - \tab 4 \tab 14 \tab 6 \tab 25 \tab - \tab 2 \tab 6 \tab 15 \tab 19 \tab 3 \tab 2 \tab 33 \tab 2 \tab 44 \tab 4 \tab 2 \tab 2 \tab 1 \tab 18 \tab 1 \tab 1 \tab 1 \tab 5 \tab 18 \tab 8 \tab 6 \tab - \tab 9 \tab 2 \tab 2 \tab 6 \tab 12 \tab 5 \tab 1 \tab 1 \tab 13 \tab 2 \tab 2 \tab 4 \tab 1 \tab 2 \tab 3 \tab 4 \tab - \tab 7 \tab - \tab 3 \tab - \tab 1 \tab 3 \tab - \tab 22 \tab 8 \tab - \tab 5 \tab - \tab 231 \tab 8 \tab 10 \tab 4 \tab 5 \tab 1 \tab 5 \tab - \tab 2 \tab 3 \tab 2 \tab 4 \tab 2 \tab 2 \tab 8 \tab - \tab 6 \tab - \tab 2 \tab 12 \tab 26 \tab - \tab 16 \tab 4 \tab - \tab 2 \tab 4 \tab 6 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 4 \tab 4 \tab - \tab - \tab - \tab 2 \tab 2 \tab 8 \tab 1 \tab - \tab 19 \tab 2 \tab 1 \tab 8 \tab 3 \tab 19 \tab 9 \tab - \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 4 \tab 1 \tab 1 \tab 8 \tab 7 \tab 4 \tab - \tab 6 \tab 4 \tab - \tab 3 \tab 2 \tab 4 \tab 11 \tab 5 \tab 5 \tab 2 \tab 5 \tab - \tab 1 \tab - \tab - \tab 1 \tab 8 \tab 7 \tab 2 \tab 8 \tab 294 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 4 \tab 3 \tab 1 \tab - \tab 3 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 6 \tab - \tab - \tab 1 \tab 2 \tab 10 \tab - \tab 25 \tab 3 \tab - \tab 7 \tab 12 \tab - \tab - \tab 8 \tab 2 \tab - \tab 2 \tab 3 \tab 3 \tab 8 \tab 1 \tab - \tab 18 \tab - \tab 1 \tab 3 \tab 10 \tab 12 \tab 14 \tab - \tab 3 \tab 12 \tab 3 \tab 26 \tab - \tab 2 \tab 9 \tab 17 \tab 14 \tab 4 \tab 5 \tab 56 \tab - \tab 46 \tab 2 \tab 4 \tab 3 \tab 1 \tab 14 \tab 1 \tab 1 \tab 4 \tab 1 \tab 20 \tab 1 \tab 5 \tab 2 \tab 10 \tab 2 \tab 1 \tab 1 \tab 17 \tab 4 \tab 4 \tab - \tab 17 \tab 2 \tab 1 \tab 5 \tab 2 \tab 5 \tab - \tab 6 \tab 1 \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab 26 \tab 9 \tab - \tab 8 \tab - \tab 263 \tab 5 \tab 11 \tab 6 \tab 1 \tab 1 \tab 4 \tab - \tab 6 \tab 1 \tab 5 \tab 7 \tab 1 \tab - \tab 4 \tab - \tab 7 \tab - \tab 3 \tab 8 \tab 27 \tab 2 \tab 14 \tab 5 \tab - \tab - \tab 7 \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 8 \tab 1 \tab - \tab 1 \tab 2 \tab 5 \tab 8 \tab 2 \tab 4 \tab 7 \tab 3 \tab 2 \tab 6 \tab 2 \tab 24 \tab 8 \tab - \tab 10 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 9 \tab - \tab 4 \tab 5 \tab 2 \tab - \tab - \tab 8 \tab 2 \tab - \tab 2 \tab 1 \tab 6 \tab 7 \tab 5 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 6 \tab - \tab 9 \tab 289 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 4 \tab - \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 9 \tab - \tab 42 \tab - \tab 1 \tab 7 \tab 8 \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 13 \tab 4 \tab 1 \tab 24 \tab 1 \tab - \tab 5 \tab 7 \tab 12 \tab 11 \tab - \tab 8 \tab 19 \tab 8 \tab 27 \tab 1 \tab 6 \tab 6 \tab 28 \tab 24 \tab 6 \tab 6 \tab 47 \tab 3 \tab 39 \tab 2 \tab 1 \tab 5 \tab 2 \tab 4 \tab 1 \tab 1 \tab 2 \tab 2 \tab 26 \tab 7 \tab 3 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab 20 \tab 3 \tab 5 \tab - \tab 21 \tab - \tab - \tab 1 \tab 3 \tab 5 \tab 6 \tab 4 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 30 \tab 6 \tab - \tab 9 \tab - \tab 252 \tab 10 \tab 12 \tab 4 \tab 4 \tab 2 \tab 6 \tab - \tab - \tab - \tab 8 \tab 4 \tab 3 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab 2 \tab 7 \tab 39 \tab 2 \tab 8 \tab 7 \tab - \tab 2 \tab 4 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 7 \tab 8 \tab - \tab - \tab - \tab 2 \tab 3 \tab 5 \tab - \tab 1 \tab 16 \tab - \tab 2 \tab 11 \tab 4 \tab 22 \tab 9 \tab 2 \tab 9 \tab - \tab 1 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab 4 \tab - \tab - \tab 8 \tab 6 \tab - \tab - \tab 9 \tab 3 \tab - \tab 4 \tab 1 \tab 5 \tab 7 \tab 6 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 6 \tab 5 \tab 5 \tab 1 \tab 12 \tab 262 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 2 \tab - \tab 5 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 11 \tab 3 \tab 43 \tab 2 \tab - \tab 8 \tab 8 \tab 1 \tab - \tab 2 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab 7 \tab 1 \tab - \tab 17 \tab 3 \tab - \tab 9 \tab 14 \tab 11 \tab 24 \tab - \tab 5 \tab 17 \tab 3 \tab 45 \tab 1 \tab 1 \tab 6 \tab 26 \tab 14 \tab 1 \tab 4 \tab 36 \tab 3 \tab 30 \tab 3 \tab 4 \tab 4 \tab 1 \tab 8 \tab 4 \tab 2 \tab 3 \tab 8 \tab 17 \tab 4 \tab 4 \tab 1 \tab 3 \tab 6 \tab - \tab 2 \tab 18 \tab 2 \tab 4 \tab - \tab 8 \tab 1 \tab - \tab 2 \tab - \tab 5 \tab 5 \tab 10 \tab - \tab 5 \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 32 \tab 5 \tab - \tab 13 \tab 1 \tab 251 \tab 11 \tab 9 \tab 5 \tab 4 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 5 \tab 1 \tab - \tab 8 \tab - \tab 6 \tab 1 \tab 4 \tab 9 \tab 13 \tab - \tab 11 \tab 8 \tab - \tab 3 \tab 5 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 4 \tab 8 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 13 \tab 2 \tab 2 \tab 4 \tab 2 \tab - \tab 7 \tab 3 \tab 13 \tab 12 \tab - \tab 1 \tab - \tab - \tab 5 \tab - \tab - \tab 1 \tab 3 \tab 6 \tab - \tab 1 \tab 4 \tab 1 \tab 3 \tab - \tab 8 \tab 3 \tab - \tab 6 \tab - \tab 1 \tab 7 \tab 2 \tab 4 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab 4 \tab 9 \tab - \tab 7 \tab 365 \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 6 \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 3 \tab - \tab 1 \tab 3 \tab 3 \tab 16 \tab 2 \tab 41 \tab 1 \tab 1 \tab 11 \tab 6 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 1 \tab 1 \tab - \tab 13 \tab 2 \tab 1 \tab 11 \tab 1 \tab 3 \tab 3 \tab 15 \tab 9 \tab 15 \tab 1 \tab 4 \tab 14 \tab 2 \tab 39 \tab 3 \tab 4 \tab 7 \tab 18 \tab 26 \tab 2 \tab 11 \tab 41 \tab 1 \tab 56 \tab 1 \tab 4 \tab 4 \tab 2 \tab 8 \tab 1 \tab 2 \tab 1 \tab 6 \tab 19 \tab 5 \tab 8 \tab 5 \tab 4 \tab 2 \tab 2 \tab 4 \tab 22 \tab 2 \tab 2 \tab - \tab 22 \tab 3 \tab 2 \tab - \tab 5 \tab 5 \tab 11 \tab 6 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 10 \tab 29 \tab 9 \tab - \tab 13 \tab - \tab 217 \tab 4 \tab 8 \tab 7 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 2 \tab 7 \tab - \tab 4 \tab 1 \tab 5 \tab 7 \tab 36 \tab - \tab 14 \tab 6 \tab 1 \tab 5 \tab 6 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 6 \tab 11 \tab 1 \tab - \tab 1 \tab 2 \tab 11 \tab 7 \tab 4 \tab 7 \tab 24 \tab 3 \tab - \tab 8 \tab 2 \tab 27 \tab 15 \tab - \tab 2 \tab - \tab - \tab 4 \tab - \tab - \tab 2 \tab 3 \tab 4 \tab - \tab 3 \tab 5 \tab 4 \tab 1 \tab 1 \tab 6 \tab 3 \tab 2 \tab 1 \tab - \tab 7 \tab 10 \tab 6 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 5 \tab 12 \tab - \tab 7 \tab 220 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 11 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 5 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 7 \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 35 \tab 1 \tab 1 \tab 10 \tab 8 \tab - \tab - \tab 6 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 10 \tab 3 \tab - \tab 14 \tab 2 \tab - \tab 5 \tab 9 \tab 8 \tab 12 \tab - \tab 1 \tab 14 \tab 8 \tab 33 \tab 1 \tab 1 \tab 5 \tab 18 \tab 13 \tab 3 \tab 6 \tab 36 \tab 1 \tab 38 \tab - \tab 3 \tab 3 \tab 2 \tab 14 \tab 3 \tab 1 \tab 2 \tab 2 \tab 17 \tab 8 \tab 7 \tab 2 \tab 9 \tab - \tab 1 \tab 5 \tab 16 \tab 2 \tab 2 \tab - \tab 19 \tab 2 \tab 1 \tab 3 \tab 4 \tab 3 \tab 7 \tab 2 \tab 1 \tab 3 \tab - \tab 5 \tab - \tab - \tab 1 \tab 1 \tab 25 \tab 12 \tab - \tab 11 \tab - \tab 204 \tab 7 \tab 4 \tab 4 \tab 7 \tab 4 \tab 4 \tab - \tab - \tab - \tab 6 \tab 4 \tab - \tab 2 \tab 4 \tab - \tab 2 \tab 1 \tab - \tab 11 \tab 46 \tab 1 \tab 14 \tab 3 \tab 1 \tab - \tab 7 \tab 4 \tab 1 \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab 10 \tab 12 \tab - \tab - \tab 5 \tab 3 \tab 6 \tab 11 \tab 2 \tab 5 \tab 14 \tab - \tab 1 \tab 7 \tab 5 \tab 18 \tab 11 \tab 2 \tab 6 \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 3 \tab 7 \tab 2 \tab - \tab 4 \tab 5 \tab 1 \tab 3 \tab 5 \tab 5 \tab 1 \tab 7 \tab 1 \tab 3 \tab 9 \tab 5 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 4 \tab 12 \tab 13 \tab - \tab 9 \tab 282 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 6 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 5 \tab 2 \tab - \tab - \tab - \tab 12 \tab 5 \tab 57 \tab 3 \tab 1 \tab 19 \tab 8 \tab - \tab - \tab 9 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 11 \tab 1 \tab 5 \tab 24 \tab - \tab 4 \tab 7 \tab 13 \tab 10 \tab 13 \tab 1 \tab 3 \tab 29 \tab 5 \tab 34 \tab - \tab 5 \tab 4 \tab 35 \tab 29 \tab 4 \tab 9 \tab 63 \tab 3 \tab 46 \tab - \tab - \tab 1 \tab 1 \tab 16 \tab 7 \tab 2 \tab 7 \tab 7 \tab 19 \tab 4 \tab 2 \tab 3 \tab 7 \tab 3 \tab 2 \tab 7 \tab 18 \tab 4 \tab 4 \tab 1 \tab 12 \tab 2 \tab 1 \tab 10 \tab 2 \tab 9 \tab 2 \tab 8 \tab 1 \tab 12 \tab - \tab 4 \tab - \tab 2 \tab 2 \tab 5 \tab 30 \tab 11 \tab - \tab 12 \tab 2 \tab 247 \tab 2 \tab 4 \tab 7 \tab 4 \tab 4 \tab 9 \tab 1 \tab 2 \tab 1 \tab 1 \tab 8 \tab 2 \tab - \tab 11 \tab - \tab 3 \tab 1 \tab 4 \tab 10 \tab 34 \tab - \tab 19 \tab 7 \tab - \tab - \tab 16 \tab - \tab - \tab - \tab 1 \tab 2 \tab 4 \tab 2 \tab - \tab 5 \tab 7 \tab - \tab - \tab 1 \tab 2 \tab 8 \tab 10 \tab 2 \tab 9 \tab 30 \tab 1 \tab 1 \tab 10 \tab 2 \tab 22 \tab 4 \tab 1 \tab 8 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 3 \tab 1 \tab 4 \tab - \tab 1 \tab 9 \tab 4 \tab 2 \tab - \tab 12 \tab 6 \tab - \tab 9 \tab - \tab 3 \tab 7 \tab 7 \tab 8 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 7 \tab 3 \tab 8 \tab - \tab 8 \tab 382 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 6 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab 7 \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 6 \tab 3 \tab 41 \tab 2 \tab 1 \tab 14 \tab 7 \tab - \tab 1 \tab 9 \tab 1 \tab 3 \tab 3 \tab 2 \tab 3 \tab 10 \tab 7 \tab 1 \tab 14 \tab 1 \tab - \tab 6 \tab 5 \tab 10 \tab 15 \tab 3 \tab 2 \tab 22 \tab 9 \tab 24 \tab 1 \tab 1 \tab 8 \tab 19 \tab 18 \tab 2 \tab 10 \tab 53 \tab 4 \tab 49 \tab 1 \tab 6 \tab 4 \tab 2 \tab 8 \tab 4 \tab 1 \tab 5 \tab 3 \tab 26 \tab 2 \tab 5 \tab 2 \tab 9 \tab 2 \tab 7 \tab 6 \tab 21 \tab 4 \tab 2 \tab - \tab 23 \tab 1 \tab - \tab 4 \tab 5 \tab 6 \tab 9 \tab 5 \tab 1 \tab 6 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 4 \tab 44 \tab 7 \tab - \tab 15 \tab 1 \tab 176 \tab 3 \tab 8 \tab 2 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab 2 \tab 6 \tab 2 \tab - \tab 13 \tab - \tab 4 \tab - \tab 2 \tab 7 \tab 40 \tab - \tab 19 \tab 8 \tab - \tab 1 \tab 11 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 3 \tab 1 \tab 8 \tab 10 \tab 1 \tab - \tab 3 \tab 6 \tab 6 \tab 6 \tab 1 \tab 4 \tab 15 \tab 1 \tab 1 \tab 8 \tab 7 \tab 33 \tab 8 \tab - \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 6 \tab - \tab 6 \tab 1 \tab 1 \tab 10 \tab 4 \tab 6 \tab - \tab 12 \tab 3 \tab 1 \tab 5 \tab 1 \tab 14 \tab 10 \tab 7 \tab 6 \tab 5 \tab - \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 12 \tab 11 \tab - \tab 17 \tab 240 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 5 \tab 4 \tab - \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 5 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1 \tab - \tab - \tab - \tab - \tab 7 \tab 2 \tab 34 \tab - \tab 3 \tab 11 \tab 3 \tab 2 \tab 1 \tab 9 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 15 \tab 5 \tab 1 \tab 16 \tab 4 \tab 3 \tab 4 \tab 6 \tab 8 \tab 14 \tab - \tab 8 \tab 15 \tab 3 \tab 22 \tab 1 \tab - \tab 6 \tab 21 \tab 22 \tab 4 \tab 9 \tab 46 \tab 1 \tab 29 \tab 2 \tab - \tab 1 \tab 5 \tab 17 \tab 3 \tab 2 \tab 5 \tab 1 \tab 15 \tab 3 \tab 2 \tab 2 \tab 7 \tab 4 \tab 5 \tab 7 \tab 8 \tab 2 \tab 5 \tab - \tab 21 \tab 3 \tab - \tab 6 \tab 1 \tab 3 \tab 8 \tab 6 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab 31 \tab 8 \tab - \tab 15 \tab 1 \tab 220 \tab 1 \tab 6 \tab 3 \tab 3 \tab - \tab 6 \tab 1 \tab 2 \tab - \tab 4 \tab 3 \tab 1 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 5 \tab 11 \tab 45 \tab 1 \tab 13 \tab 3 \tab - \tab 1 \tab 11 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 7 \tab 10 \tab 1 \tab - \tab - \tab - \tab 11 \tab 11 \tab 4 \tab 4 \tab 19 \tab - \tab 2 \tab 10 \tab 2 \tab 28 \tab 3 \tab 1 \tab 6 \tab - \tab - \tab 4 \tab 4 \tab - \tab 2 \tab 4 \tab 6 \tab 1 \tab 4 \tab 13 \tab 3 \tab 2 \tab - \tab 6 \tab 7 \tab - \tab 4 \tab - \tab 2 \tab 11 \tab 9 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 9 \tab 7 \tab 2 \tab 5 \tab 221 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 3 \tab - \tab 8 \tab 2 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 5 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 4 \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab 5 \tab 37 \tab - \tab 1 \tab 14 \tab 5 \tab 1 \tab - \tab 14 \tab - \tab 4 \tab 2 \tab 3 \tab 4 \tab 14 \tab 1 \tab 2 \tab 21 \tab 2 \tab 1 \tab 4 \tab 2 \tab 6 \tab 14 \tab 1 \tab 3 \tab 13 \tab 3 \tab 16 \tab - \tab 2 \tab 5 \tab 16 \tab 22 \tab 1 \tab 7 \tab 46 \tab 6 \tab 41 \tab - \tab 4 \tab 6 \tab 3 \tab 10 \tab 2 \tab - \tab 2 \tab 9 \tab 23 \tab 3 \tab 2 \tab - \tab 5 \tab 1 \tab 1 \tab 3 \tab 12 \tab - \tab 8 \tab 2 \tab 21 \tab 2 \tab - \tab - \tab 4 \tab 1 \tab 6 \tab 2 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 7 \tab 40 \tab 8 \tab - \tab 14 \tab - \tab 217 \tab 1 \tab 8 \tab 8 \tab 3 \tab 2 \tab 5 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 9 \tab - \tab 5 \tab 1 \tab 4 \tab 9 \tab 42 \tab 1 \tab 17 \tab 5 \tab - \tab 1 \tab 9 \tab 2 \tab - \tab - \tab 1 \tab 6 \tab 4 \tab 1 \tab - \tab 6 \tab 11 \tab - \tab 2 \tab 1 \tab - \tab 6 \tab 5 \tab - \tab 5 \tab 14 \tab 1 \tab - \tab 10 \tab 9 \tab 26 \tab 11 \tab 2 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 7 \tab - \tab 3 \tab 11 \tab 5 \tab 3 \tab - \tab 13 \tab 2 \tab 3 \tab 9 \tab 1 \tab 8 \tab 10 \tab 5 \tab 7 \tab 3 \tab 6 \tab - \tab - \tab - \tab - \tab 4 \tab 9 \tab 11 \tab 2 \tab 13 \tab 248 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 4 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 8 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 4 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 4 \tab 42 \tab 1 \tab 2 \tab 4 \tab 7 \tab 1 \tab - \tab 3 \tab 2 \tab 2 \tab 4 \tab 1 \tab - \tab 15 \tab 4 \tab - \tab 15 \tab - \tab 1 \tab 6 \tab 6 \tab 8 \tab 17 \tab - \tab 2 \tab 15 \tab 5 \tab 11 \tab - \tab 4 \tab 3 \tab 17 \tab 23 \tab 2 \tab 5 \tab 35 \tab - \tab 37 \tab 2 \tab 6 \tab 3 \tab 3 \tab 13 \tab 4 \tab 2 \tab - \tab 2 \tab 18 \tab 3 \tab 3 \tab 5 \tab 4 \tab 1 \tab 2 \tab 7 \tab 23 \tab 3 \tab 1 \tab - \tab 16 \tab 3 \tab - \tab 4 \tab 4 \tab 7 \tab 6 \tab 5 \tab - \tab 9 \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab 8 \tab 32 \tab 9 \tab - \tab 14 \tab 1 \tab 193 \tab - \tab 4 \tab 5 \tab 2 \tab 2 \tab 7 \tab - \tab 3 \tab 2 \tab 3 \tab 5 \tab 3 \tab - \tab 6 \tab - \tab 4 \tab - \tab 5 \tab 8 \tab 39 \tab - \tab 26 \tab 5 \tab 2 \tab 1 \tab 11 \tab 1 \tab - \tab - \tab - \tab 1 \tab 5 \tab 2 \tab - \tab 7 \tab 5 \tab - \tab - \tab 1 \tab 7 \tab 9 \tab 5 \tab - \tab 4 \tab 26 \tab 2 \tab 3 \tab 8 \tab 6 \tab 35 \tab 10 \tab - \tab 6 \tab - \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 7 \tab 5 \tab 1 \tab 1 \tab 10 \tab 8 \tab 1 \tab 1 \tab 9 \tab 2 \tab 2 \tab 4 \tab - \tab 4 \tab 4 \tab 6 \tab 5 \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 9 \tab 10 \tab 1 \tab 8 \tab 253 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab 5 \tab 5 \tab 65 \tab 1 \tab 1 \tab 14 \tab 6 \tab - \tab 1 \tab 9 \tab 2 \tab 4 \tab 2 \tab 2 \tab 1 \tab 10 \tab 2 \tab 2 \tab 22 \tab 1 \tab - \tab 4 \tab 5 \tab 13 \tab 24 \tab - \tab 5 \tab 21 \tab 6 \tab 28 \tab 1 \tab 2 \tab 3 \tab 23 \tab 33 \tab 8 \tab 7 \tab 63 \tab 3 \tab 51 \tab - \tab 3 \tab 4 \tab - \tab 12 \tab 3 \tab - \tab 3 \tab 4 \tab 18 \tab 2 \tab 1 \tab 6 \tab 2 \tab 3 \tab 3 \tab 6 \tab 13 \tab 3 \tab 1 \tab - \tab 21 \tab 2 \tab - \tab 3 \tab 9 \tab 3 \tab 7 \tab 4 \tab 1 \tab 7 \tab - \tab 6 \tab 1 \tab - \tab 2 \tab 5 \tab 33 \tab 6 \tab 1 \tab 10 \tab - \tab 251 \tab 2 \tab 2 \tab 8 \tab 2 \tab 2 \tab 6 \tab - \tab - \tab - \tab 6 \tab 6 \tab 2 \tab 2 \tab 6 \tab - \tab 3 \tab - \tab 1 \tab 8 \tab 32 \tab - \tab 12 \tab 3 \tab 1 \tab 2 \tab 9 \tab 5 \tab - \tab - \tab 1 \tab 3 \tab 4 \tab 1 \tab - \tab 7 \tab 13 \tab - \tab - \tab - \tab 1 \tab 5 \tab 12 \tab - \tab 8 \tab 14 \tab 2 \tab - \tab 12 \tab 4 \tab 32 \tab 7 \tab - \tab 4 \tab - \tab - \tab 4 \tab - \tab - \tab 2 \tab - \tab 6 \tab - \tab 1 \tab 9 \tab 4 \tab 1 \tab 1 \tab 11 \tab 3 \tab 1 \tab 7 \tab - \tab 7 \tab 15 \tab 10 \tab 7 \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab 12 \tab 9 \tab 13 \tab - \tab 15 \tab 382 \tab 2 \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 6 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 6 \tab 5 \tab - \tab 2 \tab 2 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 3 \tab - \tab - \tab 2 \tab - \tab 9 \tab 1 \tab 37 \tab - \tab 1 \tab 5 \tab 3 \tab - \tab - \tab 6 \tab - \tab 1 \tab 1 \tab 4 \tab 2 \tab 12 \tab - \tab 1 \tab 15 \tab - \tab 3 \tab 5 \tab 6 \tab 7 \tab 12 \tab 1 \tab 4 \tab 18 \tab 1 \tab 11 \tab - \tab 3 \tab 6 \tab 18 \tab 28 \tab 2 \tab 4 \tab 47 \tab 2 \tab 33 \tab - \tab 4 \tab - \tab 3 \tab 9 \tab 3 \tab 1 \tab 1 \tab 4 \tab 15 \tab 8 \tab 3 \tab 4 \tab 3 \tab 3 \tab 3 \tab 5 \tab 18 \tab 3 \tab 1 \tab - \tab 23 \tab 3 \tab - \tab 7 \tab 3 \tab 5 \tab 8 \tab 3 \tab 1 \tab 4 \tab - \tab 7 \tab - \tab 4 \tab 1 \tab 6 \tab 24 \tab 5 \tab - \tab 12 \tab - \tab 185 \tab 3 \tab 6 \tab 4 \tab 5 \tab - \tab 3 \tab - \tab - \tab 2 \tab 3 \tab 4 \tab 1 \tab 1 \tab 6 \tab - \tab 1 \tab - \tab 3 \tab 10 \tab 58 \tab - \tab 22 \tab 5 \tab 1 \tab 1 \tab 12 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 6 \tab 9 \tab - \tab - \tab 1 \tab 4 \tab 10 \tab 8 \tab 3 \tab 2 \tab 21 \tab 1 \tab 2 \tab 5 \tab 8 \tab 29 \tab 10 \tab - \tab 2 \tab - \tab - \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 3 \tab - \tab 14 \tab 5 \tab - \tab - \tab 10 \tab 1 \tab - \tab 7 \tab - \tab 7 \tab 11 \tab 7 \tab 5 \tab 2 \tab 3 \tab - \tab 1 \tab 5 \tab - \tab 6 \tab 11 \tab 9 \tab 1 \tab 10 \tab 227 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 5 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 9 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 7 \tab - \tab - \tab 4 \tab 2 \tab 7 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 4 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab 6 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 5 \tab 1 \tab 45 \tab 1 \tab 3 \tab 12 \tab 7 \tab 3 \tab - \tab 5 \tab 4 \tab 5 \tab - \tab 1 \tab 3 \tab 10 \tab 4 \tab 1 \tab 21 \tab 1 \tab - \tab 4 \tab 2 \tab 6 \tab 9 \tab - \tab 8 \tab 8 \tab 3 \tab 20 \tab 1 \tab 1 \tab 4 \tab 17 \tab 24 \tab - \tab 11 \tab 64 \tab 3 \tab 24 \tab 1 \tab - \tab 3 \tab 1 \tab 13 \tab 2 \tab 5 \tab - \tab 3 \tab 29 \tab 6 \tab 6 \tab 1 \tab 6 \tab 4 \tab 2 \tab 4 \tab 21 \tab - \tab 3 \tab 1 \tab 20 \tab - \tab 1 \tab 5 \tab 4 \tab 1 \tab 6 \tab 3 \tab 1 \tab 8 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 9 \tab 36 \tab 7 \tab 1 \tab 20 \tab - \tab 195 \tab - \tab 6 \tab 7 \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab 4 \tab 3 \tab 1 \tab 1 \tab - \tab 9 \tab - \tab 6 \tab - \tab 3 \tab 6 \tab 32 \tab - \tab 14 \tab 2 \tab 1 \tab 1 \tab 9 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 5 \tab 13 \tab 1 \tab - \tab 3 \tab 4 \tab 10 \tab 11 \tab 3 \tab 7 \tab 16 \tab 2 \tab 3 \tab 8 \tab 6 \tab 27 \tab 9 \tab 1 \tab 13 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 7 \tab 1 \tab 1 \tab 12 \tab 2 \tab 2 \tab - \tab 12 \tab 5 \tab 2 \tab 9 \tab 1 \tab 9 \tab 8 \tab 4 \tab 8 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 9 \tab 14 \tab 11 \tab - \tab 7 \tab 247 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 3 \tab - \tab 10 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab - \tab - \tab 3 \tab - \tab 7 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 10 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 6 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 4 \tab - \tab 26 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 12 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 10 \tab 2 \tab 1 \tab 20 \tab 2 \tab 2 \tab 10 \tab 2 \tab 11 \tab 23 \tab - \tab 9 \tab 11 \tab 5 \tab 11 \tab 1 \tab 2 \tab 3 \tab 21 \tab 26 \tab 1 \tab 9 \tab 44 \tab 1 \tab 32 \tab 3 \tab 5 \tab 2 \tab 1 \tab 12 \tab 2 \tab 2 \tab 1 \tab 3 \tab 21 \tab 3 \tab 2 \tab 4 \tab 2 \tab 3 \tab 1 \tab 2 \tab 6 \tab 1 \tab 3 \tab 2 \tab 8 \tab - \tab - \tab 3 \tab 2 \tab 3 \tab 12 \tab 2 \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 12 \tab 23 \tab 5 \tab - \tab 5 \tab - \tab 151 \tab 2 \tab 3 \tab 3 \tab 6 \tab 1 \tab 5 \tab 1 \tab - \tab 1 \tab 4 \tab 5 \tab 2 \tab - \tab 6 \tab 1 \tab 3 \tab - \tab 1 \tab 8 \tab 44 \tab - \tab 19 \tab 10 \tab 1 \tab 5 \tab 8 \tab 6 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 5 \tab 24 \tab - \tab - \tab 1 \tab 1 \tab 11 \tab 7 \tab 5 \tab 3 \tab 17 \tab 1 \tab 1 \tab 7 \tab 6 \tab 23 \tab 5 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab - \tab 3 \tab 3 \tab 5 \tab 3 \tab - \tab 10 \tab 2 \tab 1 \tab 1 \tab 8 \tab 1 \tab 1 \tab 5 \tab 1 \tab 8 \tab 7 \tab 8 \tab 6 \tab 5 \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 9 \tab 4 \tab 1 \tab 9 \tab 233 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab 2 \tab - \tab 4 \tab - \tab 4 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 5 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab 5 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 6 \tab - \tab - \tab 1 \tab - \tab 3 \tab 4 \tab 39 \tab - \tab 2 \tab 7 \tab 5 \tab 1 \tab - \tab 4 \tab 3 \tab 1 \tab 3 \tab - \tab 2 \tab 12 \tab 2 \tab 1 \tab 18 \tab 1 \tab 2 \tab 8 \tab 2 \tab 11 \tab 15 \tab - \tab 1 \tab 20 \tab 5 \tab 15 \tab 1 \tab 4 \tab 5 \tab 19 \tab 21 \tab - \tab 12 \tab 51 \tab 2 \tab 34 \tab 2 \tab 1 \tab 2 \tab 1 \tab 14 \tab - \tab 2 \tab 1 \tab 3 \tab 22 \tab 2 \tab 7 \tab 1 \tab 6 \tab 1 \tab - \tab 2 \tab 10 \tab 4 \tab 8 \tab 1 \tab 20 \tab 3 \tab - \tab 4 \tab 6 \tab 3 \tab 6 \tab 1 \tab - \tab 5 \tab - \tab 2 \tab - \tab - \tab - \tab 8 \tab 24 \tab 4 \tab - \tab 16 \tab 1 \tab 173 \tab 1 \tab 4 \tab 2 \tab 3 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 11 \tab - \tab 6 \tab - \tab 4 \tab 10 \tab 53 \tab - \tab 12 \tab 8 \tab 1 \tab - \tab 12 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab 13 \tab 11 \tab - \tab - \tab - \tab 3 \tab 7 \tab 6 \tab 4 \tab 1 \tab 29 \tab 1 \tab - \tab 6 \tab 2 \tab 33 \tab 2 \tab - \tab 5 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab 5 \tab 2 \tab - \tab 7 \tab 4 \tab 1 \tab 1 \tab 8 \tab 5 \tab 1 \tab 10 \tab 1 \tab 7 \tab 12 \tab 11 \tab 6 \tab 4 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab 7 \tab 7 \tab 5 \tab - \tab 14 \tab 266 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 6 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 6 \tab - \tab 6 \tab 2 \tab 2 \tab - \tab - \tab 11 \tab 2 \tab 1 \tab 1 \tab 2 \tab 7 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 5 \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 6 \tab - \tab - \tab 1 \tab 10 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 6 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 3 \tab 3 \tab - \tab - \tab 2 \tab 9 \tab 2 \tab 42 \tab - \tab 1 \tab 9 \tab 8 \tab - \tab - \tab 6 \tab - \tab 2 \tab 2 \tab 1 \tab 3 \tab 12 \tab 2 \tab 1 \tab 10 \tab - \tab 4 \tab 8 \tab 5 \tab 7 \tab 14 \tab 2 \tab 3 \tab 17 \tab 5 \tab 4 \tab 1 \tab 3 \tab 5 \tab 24 \tab 25 \tab 2 \tab 11 \tab 49 \tab 1 \tab 35 \tab - \tab 2 \tab 1 \tab 3 \tab 10 \tab 9 \tab 1 \tab 3 \tab 3 \tab 23 \tab 1 \tab 6 \tab 2 \tab 3 \tab 1 \tab 5 \tab 3 \tab 14 \tab - \tab 4 \tab - \tab 26 \tab 2 \tab - \tab 3 \tab 4 \tab 1 \tab 7 \tab 9 \tab 2 \tab 3 \tab - \tab 4 \tab - \tab 2 \tab - \tab 13 \tab 30 \tab 11 \tab - \tab 6 \tab 1 \tab 189 \tab 2 \tab 6 \tab 4 \tab 3 \tab 2 \tab 4 \tab - \tab 1 \tab - \tab 7 \tab 3 \tab 2 \tab 1 \tab 8 \tab - \tab 4 \tab - \tab 1 \tab 14 \tab 43 \tab - \tab 24 \tab 7 \tab 1 \tab 2 \tab 16 \tab 6 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 10 \tab 13 \tab 1 \tab 1 \tab - \tab 3 \tab 8 \tab 7 \tab - \tab 9 \tab 23 \tab 1 \tab 1 \tab 12 \tab 4 \tab 33 \tab 15 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab 9 \tab 1 \tab - \tab 15 \tab 13 \tab 1 \tab - \tab 10 \tab 3 \tab - \tab 13 \tab 2 \tab 5 \tab 11 \tab 6 \tab 10 \tab 2 \tab 1 \tab - \tab - \tab 4 \tab - \tab 9 \tab 14 \tab 13 \tab - \tab 18 \tab 244 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 10 \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab 6 \tab 1 \tab - \tab 2 \tab - \tab 8 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 4 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 6 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 5 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 8 \tab 2 \tab - \tab - \tab 6 \tab 12 \tab 4 \tab 58 \tab - \tab 2 \tab 13 \tab 9 \tab 1 \tab 1 \tab 13 \tab 2 \tab - \tab 6 \tab 1 \tab - \tab 21 \tab 3 \tab 3 \tab 39 \tab 3 \tab 5 \tab 14 \tab 7 \tab 16 \tab 34 \tab - \tab 6 \tab 50 \tab 11 \tab 15 \tab - \tab 5 \tab 10 \tab 48 \tab 42 \tab 9 \tab 25 \tab 107 \tab 5 \tab 55 \tab 4 \tab 5 \tab 4 \tab 5 \tab 18 \tab 7 \tab 4 \tab 6 \tab 5 \tab 38 \tab 9 \tab 5 \tab 6 \tab 10 \tab 1 \tab 5 \tab 7 \tab 36 \tab 2 \tab 8 \tab 2 \tab 30 \tab 4 \tab 1 \tab 4 \tab 9 \tab 5 \tab 19 \tab 3 \tab - \tab 10 \tab 2 \tab 9 \tab 2 \tab - \tab 1 \tab 9 \tab 46 \tab 14 \tab - \tab 23 \tab 2 \tab 290 \tab 8 \tab 7 \tab 14 \tab 8 \tab 1 \tab 9 \tab - \tab 1 \tab 1 \tab 8 \tab 9 \tab 5 \tab 3 \tab 14 \tab 1 \tab 3 \tab - \tab 5 \tab 23 \tab 69 \tab 1 \tab 26 \tab 13 \tab 1 \tab 7 \tab 20 \tab 10 \tab - \tab - \tab 4 \tab 2 \tab 7 \tab 8 \tab - \tab 10 \tab 20 \tab 2 \tab - \tab 3 \tab 8 \tab 15 \tab 9 \tab 3 \tab 12 \tab 39 \tab 3 \tab 3 \tab 18 \tab 9 \tab 82 \tab 14 \tab 2 \tab 7 \tab - \tab - \tab 5 \tab 6 \tab - \tab 2 \tab 4 \tab 5 \tab 1 \tab 2 \tab 32 \tab 17 \tab 3 \tab 2 \tab 22 \tab 3 \tab 1 \tab 19 \tab 3 \tab 13 \tab 22 \tab 15 \tab 17 \tab 10 \tab 13 \tab 3 \tab 1 \tab 4 \tab 4 \tab 13 \tab 22 \tab 33 \tab 1 \tab 18 \tab 551 \tab 3 \tab 4 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 5 \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 16 \tab 2 \tab 6 \tab 4 \tab - \tab 1 \tab 1 \tab 8 \tab 3 \tab - \tab 5 \tab - \tab 13 \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 4 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 5 \tab 2 \tab - \tab - \tab 11 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 9 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab 5 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab - \tab 5 \tab 6 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 4 \tab - \tab - \tab 2 \tab 4 \tab 12 \tab 5 \tab 55 \tab 3 \tab 3 \tab 11 \tab 9 \tab - \tab 3 \tab 11 \tab 2 \tab 7 \tab 3 \tab 3 \tab 6 \tab 26 \tab 5 \tab 2 \tab 47 \tab 2 \tab 3 \tab 10 \tab 7 \tab 12 \tab 34 \tab 1 \tab 5 \tab 32 \tab 7 \tab 14 \tab - \tab 3 \tab 3 \tab 37 \tab 50 \tab 3 \tab 19 \tab 90 \tab 3 \tab 44 \tab 3 \tab 3 \tab 2 \tab 5 \tab 20 \tab 7 \tab 5 \tab 6 \tab 7 \tab 41 \tab 6 \tab 7 \tab - \tab 9 \tab 1 \tab 5 \tab 7 \tab 29 \tab 2 \tab 3 \tab 3 \tab 43 \tab 4 \tab - \tab 4 \tab 6 \tab 4 \tab 26 \tab 4 \tab 2 \tab 8 \tab 1 \tab 7 \tab 1 \tab 1 \tab 3 \tab 8 \tab 58 \tab 10 \tab - \tab 32 \tab - \tab 237 \tab 3 \tab 5 \tab 7 \tab 5 \tab 2 \tab 7 \tab 1 \tab 2 \tab 1 \tab 10 \tab 3 \tab 4 \tab 3 \tab 12 \tab - \tab 4 \tab 1 \tab 3 \tab 18 \tab 70 \tab 1 \tab 38 \tab 11 \tab 2 \tab 7 \tab 20 \tab 7 \tab - \tab - \tab 1 \tab 3 \tab 11 \tab 4 \tab - \tab 8 \tab 27 \tab 1 \tab - \tab 2 \tab 7 \tab 17 \tab 5 \tab 4 \tab 8 \tab 40 \tab 2 \tab 4 \tab 14 \tab 13 \tab 71 \tab 15 \tab 3 \tab 10 \tab - \tab - \tab 5 \tab 4 \tab - \tab 5 \tab 5 \tab 13 \tab 4 \tab 2 \tab 31 \tab 13 \tab 1 \tab 3 \tab 22 \tab 5 \tab 4 \tab 24 \tab 2 \tab 11 \tab 28 \tab 26 \tab 14 \tab 9 \tab 8 \tab - \tab - \tab 6 \tab - \tab 23 \tab 16 \tab 26 \tab 1 \tab 21 \tab 568 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 7 \tab - \tab 2 \tab 2 \tab 3 \tab - \tab - \tab 18 \tab - \tab 5 \tab 3 \tab - \tab 6 \tab - \tab 9 \tab 3 \tab 2 \tab 3 \tab - \tab 8 \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab 5 \tab 8 \tab 2 \tab 4 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 6 \tab 2 \tab - \tab 1 \tab 7 \tab 3 \tab 2 \tab - \tab - \tab - \tab 13 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 4 \tab 3 \tab 2 \tab 1 \tab 6 \tab 5 \tab 1 \tab - \tab - \tab 3 \tab - \tab 3 \tab - \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 6 \tab 1 \tab - \tab 1 \tab 2 \tab 10 \tab 6 \tab 52 \tab 2 \tab 2 \tab 21 \tab 6 \tab 1 \tab 2 \tab 15 \tab 3 \tab - \tab 5 \tab 3 \tab 4 \tab 18 \tab 3 \tab 2 \tab 25 \tab 3 \tab 4 \tab 12 \tab 6 \tab 12 \tab 18 \tab - \tab 3 \tab 36 \tab 4 \tab 22 \tab - \tab 1 \tab 4 \tab 24 \tab 49 \tab 3 \tab 29 \tab 80 \tab 3 \tab 54 \tab - \tab 2 \tab 2 \tab 3 \tab 17 \tab 9 \tab 4 \tab 6 \tab 4 \tab 40 \tab 8 \tab 7 \tab 8 \tab 3 \tab 4 \tab 2 \tab 7 \tab 40 \tab 3 \tab 6 \tab - \tab 37 \tab 2 \tab - \tab 9 \tab 7 \tab 2 \tab 20 \tab 6 \tab 2 \tab 7 \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab 10 \tab 56 \tab 20 \tab - \tab 30 \tab 2 \tab 227 \tab 2 \tab 10 \tab 5 \tab 3 \tab 2 \tab 5 \tab 1 \tab 1 \tab - \tab 10 \tab 12 \tab 1 \tab 2 \tab 16 \tab - \tab 6 \tab - \tab 2 \tab 27 \tab 73 \tab - \tab 29 \tab 10 \tab 2 \tab 3 \tab 29 \tab 7 \tab 2 \tab - \tab 6 \tab 2 \tab 4 \tab 1 \tab - \tab 14 \tab 19 \tab 1 \tab 1 \tab 2 \tab 8 \tab 14 \tab 11 \tab 6 \tab 7 \tab 25 \tab 3 \tab 1 \tab 21 \tab 13 \tab 68 \tab 22 \tab 2 \tab 12 \tab - \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 6 \tab 14 \tab - \tab 1 \tab 19 \tab 11 \tab 3 \tab 2 \tab 18 \tab 6 \tab 2 \tab 18 \tab 1 \tab 16 \tab 22 \tab 20 \tab 16 \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 21 \tab 29 \tab 29 \tab 5 \tab 24 \tab 569 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 9 \tab 1 \tab - \tab 2 \tab 5 \tab 4 \tab - \tab 8 \tab 4 \tab 10 \tab - \tab 1 \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab 6 \tab 3 \tab 15 \tab 2 \tab - \tab 1 \tab - \tab - \tab 5 \tab - \tab 4 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 2 \tab 4 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab 10 \tab 3 \tab - \tab - \tab 1 \tab - \tab 11 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 5 \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 7 \tab 8 \tab 3 \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 3 \tab 4 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 10 \tab 2 \tab 58 \tab 4 \tab 3 \tab 11 \tab 8 \tab 1 \tab 1 \tab 11 \tab 3 \tab - \tab 3 \tab 2 \tab 10 \tab 29 \tab 1 \tab 1 \tab 38 \tab 3 \tab - \tab 12 \tab 7 \tab 8 \tab 25 \tab 1 \tab 7 \tab 38 \tab 5 \tab 19 \tab 1 \tab 6 \tab 2 \tab 38 \tab 51 \tab - \tab 16 \tab 120 \tab 6 \tab 47 \tab 1 \tab 4 \tab 3 \tab 2 \tab 22 \tab 10 \tab 3 \tab 7 \tab 6 \tab 47 \tab 3 \tab 10 \tab 2 \tab 6 \tab 6 \tab - \tab 9 \tab 24 \tab 3 \tab 3 \tab - \tab 30 \tab 1 \tab - \tab 5 \tab 10 \tab 7 \tab 14 \tab 9 \tab - \tab 5 \tab - \tab 8 \tab 1 \tab 2 \tab 1 \tab 6 \tab 50 \tab 13 \tab 1 \tab 22 \tab 1 \tab 248 \tab 4 \tab 10 \tab 7 \tab 2 \tab 3 \tab 11 \tab 1 \tab - \tab 1 \tab 8 \tab 12 \tab 3 \tab 1 \tab 8 \tab - \tab 4 \tab 1 \tab 7 \tab 15 \tab 64 \tab 3 \tab 25 \tab 10 \tab 2 \tab 4 \tab 39 \tab 8 \tab - \tab - \tab 2 \tab 1 \tab 5 \tab 2 \tab - \tab 11 \tab 20 \tab 1 \tab - \tab 1 \tab 8 \tab 10 \tab 10 \tab 4 \tab 13 \tab 39 \tab 3 \tab - \tab 15 \tab 8 \tab 54 \tab 11 \tab - \tab 9 \tab 1 \tab - \tab 5 \tab 5 \tab - \tab 7 \tab 6 \tab 10 \tab - \tab 4 \tab 22 \tab 7 \tab 7 \tab 1 \tab 20 \tab 5 \tab 1 \tab 16 \tab 2 \tab 12 \tab 22 \tab 19 \tab 13 \tab 9 \tab 4 \tab - \tab 1 \tab 3 \tab - \tab 16 \tab 21 \tab 18 \tab 1 \tab 19 \tab 576 \tab 3 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 3 \tab 4 \tab 2 \tab - \tab 12 \tab 2 \tab 3 \tab 6 \tab 1 \tab 2 \tab 1 \tab 11 \tab - \tab 1 \tab 6 \tab 3 \tab 10 \tab 1 \tab - \tab - \tab - \tab 2 \tab 3 \tab 3 \tab 4 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab 7 \tab - \tab - \tab - \tab 16 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 9 \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 10 \tab 5 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 5 \tab - \tab 5 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 7 \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 13 \tab - \tab 54 \tab 2 \tab 1 \tab 11 \tab 3 \tab 1 \tab 1 \tab 11 \tab 2 \tab 1 \tab 1 \tab 4 \tab 6 \tab 16 \tab 2 \tab 1 \tab 39 \tab 2 \tab 7 \tab 12 \tab 4 \tab 8 \tab 23 \tab - \tab 8 \tab 33 \tab 7 \tab 16 \tab 1 \tab 2 \tab 7 \tab 34 \tab 66 \tab 2 \tab 12 \tab 97 \tab 1 \tab 53 \tab 1 \tab 6 \tab 2 \tab 1 \tab 16 \tab 3 \tab 1 \tab 4 \tab 5 \tab 56 \tab 4 \tab 7 \tab 6 \tab 5 \tab 2 \tab 7 \tab 5 \tab 19 \tab 4 \tab 5 \tab - \tab 34 \tab 1 \tab - \tab 4 \tab 4 \tab 4 \tab 15 \tab 8 \tab - \tab 8 \tab 2 \tab 9 \tab - \tab 2 \tab 2 \tab 13 \tab 50 \tab 11 \tab - \tab 33 \tab 1 \tab 204 \tab 4 \tab 8 \tab 9 \tab 11 \tab 5 \tab 3 \tab - \tab 3 \tab 1 \tab 5 \tab 9 \tab 3 \tab 3 \tab 16 \tab - \tab 5 \tab - \tab 5 \tab 23 \tab 64 \tab - \tab 27 \tab 11 \tab 2 \tab 8 \tab 25 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 6 \tab 2 \tab - \tab 14 \tab 24 \tab 2 \tab - \tab 3 \tab 4 \tab 15 \tab 23 \tab 8 \tab 8 \tab 32 \tab - \tab - \tab 8 \tab 11 \tab 63 \tab 24 \tab 2 \tab 9 \tab - \tab - \tab 3 \tab 4 \tab - \tab 4 \tab 1 \tab 13 \tab 2 \tab 6 \tab 18 \tab 12 \tab 4 \tab 1 \tab 16 \tab 6 \tab - \tab 16 \tab - \tab 10 \tab 26 \tab 14 \tab 15 \tab 7 \tab 8 \tab 1 \tab 1 \tab 4 \tab 1 \tab 18 \tab 21 \tab 17 \tab - \tab 21 \tab 589 \tab 5 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 7 \tab 3 \tab - \tab 1 \tab 4 \tab 2 \tab - \tab 13 \tab 4 \tab 3 \tab 2 \tab - \tab 3 \tab 1 \tab 5 \tab - \tab 1 \tab 7 \tab 3 \tab 14 \tab 2 \tab - \tab 3 \tab - \tab 4 \tab 2 \tab 3 \tab 3 \tab 5 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab 2 \tab 3 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 8 \tab 4 \tab - \tab 2 \tab 2 \tab 2 \tab 9 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab 4 \tab - \tab 2 \tab 3 \tab 3 \tab 3 \tab 2 \tab 1 \tab 5 \tab 8 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 6 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 5 \tab - \tab - \tab 1 \tab 3 \tab 14 \tab 2 \tab 62 \tab 1 \tab 1 \tab 13 \tab 2 \tab - \tab 1 \tab 15 \tab 2 \tab 1 \tab 1 \tab 5 \tab 5 \tab 19 \tab 4 \tab 4 \tab 54 \tab 3 \tab 5 \tab 13 \tab 2 \tab 18 \tab 24 \tab - \tab 9 \tab 33 \tab 5 \tab 25 \tab - \tab 5 \tab 8 \tab 26 \tab 50 \tab 1 \tab 18 \tab 93 \tab 6 \tab 55 \tab 4 \tab 2 \tab 2 \tab 2 \tab 20 \tab 4 \tab 1 \tab 7 \tab 6 \tab 38 \tab 9 \tab 7 \tab 3 \tab 4 \tab 6 \tab 1 \tab 4 \tab 32 \tab 4 \tab 6 \tab - \tab 24 \tab 4 \tab 1 \tab 10 \tab 8 \tab 4 \tab 19 \tab 5 \tab 1 \tab 7 \tab - \tab 9 \tab 1 \tab 1 \tab 5 \tab 10 \tab 57 \tab 17 \tab 1 \tab 23 \tab - \tab 172 \tab 4 \tab 16 \tab 5 \tab 9 \tab 4 \tab 6 \tab - \tab 1 \tab 3 \tab 3 \tab 8 \tab 1 \tab 2 \tab 16 \tab - \tab 4 \tab - \tab - \tab 15 \tab 56 \tab - \tab 44 \tab 10 \tab - \tab 4 \tab 29 \tab 11 \tab 1 \tab - \tab - \tab 6 \tab 6 \tab 2 \tab - \tab 8 \tab 18 \tab - \tab - \tab 2 \tab 6 \tab 16 \tab 21 \tab 5 \tab 9 \tab 27 \tab 1 \tab - \tab 15 \tab 8 \tab 60 \tab 17 \tab 2 \tab 6 \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 4 \tab 7 \tab 12 \tab 1 \tab 1 \tab 22 \tab 14 \tab 1 \tab 2 \tab 14 \tab 4 \tab - \tab 17 \tab 1 \tab 18 \tab 28 \tab 12 \tab 14 \tab 7 \tab 8 \tab 1 \tab 1 \tab 6 \tab - \tab 24 \tab 31 \tab 21 \tab - \tab 22 \tab 603 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab - \tab 11 \tab 3 \tab 4 \tab 1 \tab - \tab 4 \tab 1 \tab 5 \tab 4 \tab 1 \tab 7 \tab 1 \tab 11 \tab 3 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 14 \tab 2 \tab 2 \tab 1 \tab 5 \tab - \tab 15 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 4 \tab 1 \tab 5 \tab 6 \tab 5 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 8 \tab - \tab 1 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 10 \tab 1 \tab 1 \tab 1 \tab 5 \tab 22 \tab 3 \tab 78 \tab 4 \tab 1 \tab 16 \tab 7 \tab - \tab - \tab 15 \tab 1 \tab 3 \tab 7 \tab 7 \tab 4 \tab 24 \tab 5 \tab 3 \tab 54 \tab 4 \tab 5 \tab 13 \tab 3 \tab 15 \tab 29 \tab 1 \tab 7 \tab 38 \tab 6 \tab 13 \tab 1 \tab 3 \tab 5 \tab 34 \tab 53 \tab 1 \tab 26 \tab 102 \tab 4 \tab 55 \tab 5 \tab 7 \tab 1 \tab 1 \tab 24 \tab 5 \tab 3 \tab 9 \tab 2 \tab 43 \tab 6 \tab 6 \tab 4 \tab 6 \tab 2 \tab 3 \tab 9 \tab 24 \tab 2 \tab 2 \tab - \tab 48 \tab 1 \tab - \tab 8 \tab 12 \tab 5 \tab 15 \tab 5 \tab - \tab 7 \tab 1 \tab 3 \tab - \tab 2 \tab 5 \tab 7 \tab 40 \tab 15 \tab - \tab 34 \tab - \tab 194 \tab 1 \tab 10 \tab 5 \tab 9 \tab 5 \tab 6 \tab - \tab 2 \tab 2 \tab 11 \tab 9 \tab 1 \tab 2 \tab 14 \tab - \tab 8 \tab - \tab 10 \tab 24 \tab 66 \tab - \tab 35 \tab 5 \tab - \tab 4 \tab 25 \tab 2 \tab - \tab - \tab 5 \tab 2 \tab 6 \tab 1 \tab - \tab 18 \tab 18 \tab - \tab - \tab 1 \tab 3 \tab 11 \tab 31 \tab 7 \tab 11 \tab 32 \tab 2 \tab 1 \tab 21 \tab 17 \tab 67 \tab 16 \tab - \tab 7 \tab - \tab - \tab 5 \tab 3 \tab - \tab 3 \tab 7 \tab 16 \tab - \tab 3 \tab 23 \tab 5 \tab 3 \tab 1 \tab 23 \tab 4 \tab - \tab 18 \tab - \tab 16 \tab 27 \tab 18 \tab 15 \tab 6 \tab 9 \tab 2 \tab 1 \tab 5 \tab - \tab 10 \tab 20 \tab 30 \tab 2 \tab 28 \tab 2107 \tab 2 \tab - \tab 2 \tab - \tab - \tab 4 \tab - \tab 1 \tab 3 \tab 5 \tab - \tab - \tab 3 \tab 5 \tab 2 \tab - \tab 17 \tab - \tab 9 \tab 1 \tab - \tab 2 \tab 1 \tab 5 \tab 3 \tab 1 \tab 4 \tab - \tab 14 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 5 \tab 4 \tab 7 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 3 \tab 6 \tab - \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab 13 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 23 \tab 3 \tab 4 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 4 \tab 1 \tab - \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab 5 \tab 3 \tab 1 \tab 7 \tab 4 \tab - \tab 2 \tab - \tab 2 \tab - \tab 4 \tab - \tab - \tab 5 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 7 \tab 1 \tab 1 \tab - \tab 3 \tab 9 \tab 3 \tab 33 \tab 3 \tab 3 \tab 5 \tab 4 \tab - \tab 2 \tab 10 \tab 1 \tab 3 \tab 1 \tab 4 \tab - \tab 16 \tab 2 \tab 1 \tab 32 \tab 2 \tab 5 \tab 9 \tab 1 \tab 7 \tab 19 \tab - \tab 2 \tab 29 \tab 4 \tab 14 \tab 1 \tab 2 \tab 6 \tab 19 \tab 44 \tab 6 \tab 6 \tab 64 \tab 4 \tab 25 \tab 4 \tab 4 \tab 2 \tab 2 \tab 18 \tab 2 \tab 1 \tab 13 \tab 6 \tab 31 \tab 3 \tab 6 \tab 1 \tab 4 \tab 3 \tab 4 \tab 8 \tab 27 \tab 5 \tab 3 \tab - \tab 26 \tab - \tab - \tab 4 \tab 3 \tab 2 \tab 11 \tab 2 \tab 1 \tab 4 \tab 2 \tab 4 \tab - \tab 2 \tab 1 \tab 13 \tab 32 \tab 9 \tab - \tab 29 \tab - \tab 127 \tab 7 \tab 6 \tab 4 \tab 8 \tab 1 \tab 9 \tab - \tab - \tab - \tab 6 \tab 5 \tab 3 \tab 1 \tab 7 \tab - \tab 1 \tab - \tab 2 \tab 17 \tab 26 \tab 1 \tab 20 \tab 10 \tab 1 \tab 5 \tab 24 \tab 3 \tab - \tab - \tab 2 \tab - \tab 3 \tab 4 \tab - \tab 8 \tab 11 \tab 1 \tab - \tab 1 \tab 5 \tab 11 \tab 25 \tab 4 \tab 10 \tab 18 \tab 3 \tab - \tab 23 \tab 6 \tab 61 \tab 14 \tab 2 \tab 11 \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 3 \tab 5 \tab 5 \tab 5 \tab 1 \tab 20 \tab 8 \tab 2 \tab 2 \tab 9 \tab 4 \tab 2 \tab 11 \tab 2 \tab 16 \tab 19 \tab 18 \tab 7 \tab 6 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 8 \tab 12 \tab 12 \tab 3 \tab 13 \tab 398 \tab 2 \tab 5 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 11 \tab 2 \tab 3 \tab - \tab - \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 4 \tab 1 \tab 12 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 5 \tab - \tab - \tab - \tab 9 \tab 1 \tab - \tab - \tab 3 \tab - \tab 6 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 6 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab 5 \tab 1 \tab 2 \tab - \tab 3 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 6 \tab 4 \tab - \tab - \tab 2 \tab 7 \tab 4 \tab 39 \tab - \tab 2 \tab 6 \tab 6 \tab 2 \tab - \tab 6 \tab 3 \tab 1 \tab - \tab 4 \tab 2 \tab 22 \tab 6 \tab 2 \tab 39 \tab 2 \tab 4 \tab 9 \tab 2 \tab 8 \tab 16 \tab - \tab 7 \tab 33 \tab 7 \tab 10 \tab 2 \tab 3 \tab 4 \tab 26 \tab 25 \tab 1 \tab 13 \tab 75 \tab 3 \tab 31 \tab 1 \tab 5 \tab 3 \tab - \tab 18 \tab 4 \tab 1 \tab 8 \tab 2 \tab 33 \tab 4 \tab 3 \tab 1 \tab 5 \tab 2 \tab 3 \tab 6 \tab 16 \tab 2 \tab 1 \tab - \tab 26 \tab - \tab - \tab 6 \tab 11 \tab 5 \tab 10 \tab 7 \tab 1 \tab 4 \tab 1 \tab 10 \tab - \tab 2 \tab 2 \tab 6 \tab 33 \tab 15 \tab - \tab 19 \tab 1 \tab 114 \tab 2 \tab 12 \tab 4 \tab 5 \tab 6 \tab 6 \tab - \tab - \tab 1 \tab 3 \tab 4 \tab 2 \tab 2 \tab 5 \tab - \tab 4 \tab - \tab 2 \tab 17 \tab 50 \tab - \tab 21 \tab 4 \tab - \tab 4 \tab 16 \tab 3 \tab - \tab 1 \tab - \tab - \tab 2 \tab 5 \tab - \tab 1 \tab 15 \tab 1 \tab - \tab - \tab 5 \tab 11 \tab 18 \tab 4 \tab 4 \tab 24 \tab 3 \tab 1 \tab 10 \tab 10 \tab 60 \tab 11 \tab 2 \tab 6 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 8 \tab 13 \tab - \tab 4 \tab 22 \tab 8 \tab 4 \tab - \tab 7 \tab 5 \tab - \tab 13 \tab 1 \tab 7 \tab 15 \tab 15 \tab 11 \tab 9 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 11 \tab 24 \tab 15 \tab 1 \tab 29 \tab 402 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 14 \tab 3 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab 7 \tab 1 \tab - \tab 5 \tab 2 \tab 12 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 6 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 7 \tab - \tab 1 \tab 1 \tab - \tab - \tab 8 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 6 \tab - \tab 5 \tab 8 \tab 2 \tab - \tab - \tab 4 \tab - \tab 6 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 6 \tab 2 \tab 51 \tab - \tab 1 \tab 11 \tab 3 \tab - \tab - \tab 7 \tab 2 \tab 2 \tab 3 \tab 2 \tab 4 \tab 15 \tab 4 \tab - \tab 55 \tab 1 \tab 2 \tab 9 \tab 1 \tab 11 \tab 22 \tab 1 \tab 3 \tab 38 \tab 7 \tab 10 \tab - \tab 1 \tab 4 \tab 29 \tab 55 \tab 1 \tab 12 \tab 57 \tab 2 \tab 30 \tab 1 \tab - \tab - \tab - \tab 12 \tab 5 \tab 1 \tab 3 \tab 3 \tab 27 \tab 9 \tab 5 \tab 1 \tab 3 \tab 6 \tab 1 \tab 6 \tab 25 \tab 3 \tab 3 \tab - \tab 20 \tab - \tab - \tab 4 \tab 10 \tab - \tab 7 \tab 2 \tab 1 \tab 4 \tab - \tab 3 \tab - \tab - \tab 1 \tab 4 \tab 22 \tab 8 \tab - \tab 17 \tab - \tab 114 \tab 5 \tab 7 \tab 7 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 4 \tab 7 \tab - \tab - \tab 11 \tab - \tab 2 \tab - \tab 4 \tab 8 \tab 39 \tab 1 \tab 19 \tab 3 \tab - \tab 5 \tab 25 \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab 5 \tab 6 \tab - \tab - \tab 1 \tab 7 \tab 6 \tab - \tab - \tab 12 \tab 24 \tab 1 \tab 2 \tab 15 \tab 7 \tab 105 \tab 12 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 4 \tab 2 \tab 7 \tab 4 \tab 1 \tab 15 \tab 7 \tab 1 \tab 1 \tab 18 \tab 6 \tab 3 \tab 13 \tab 1 \tab 11 \tab 10 \tab 12 \tab 10 \tab 5 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 15 \tab 24 \tab 10 \tab 2 \tab 30 \tab 385 \tab 5 \tab 3 \tab - \tab 1 \tab 5 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 8 \tab 1 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 3 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 23 \tab 1 \tab - \tab - \tab 2 \tab - \tab 7 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 5 \tab 1 \tab 6 \tab 8 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 3 \tab 1 \tab 1 \tab 1 \tab 6 \tab 6 \tab 2 \tab 57 \tab 1 \tab 2 \tab 11 \tab 6 \tab 1 \tab 1 \tab 13 \tab 4 \tab - \tab - \tab 3 \tab 3 \tab 22 \tab 1 \tab 2 \tab 51 \tab 4 \tab 6 \tab 12 \tab 4 \tab 16 \tab 24 \tab 1 \tab 3 \tab 35 \tab 7 \tab 15 \tab 2 \tab 4 \tab 2 \tab 33 \tab 54 \tab 1 \tab 17 \tab 80 \tab 2 \tab 42 \tab 2 \tab 3 \tab 1 \tab - \tab 28 \tab 3 \tab 1 \tab 2 \tab 4 \tab 39 \tab 6 \tab 3 \tab - \tab 3 \tab 2 \tab 4 \tab 4 \tab 20 \tab 10 \tab 3 \tab - \tab 27 \tab 1 \tab - \tab 5 \tab 9 \tab 1 \tab 6 \tab 3 \tab 4 \tab 3 \tab 1 \tab 10 \tab - \tab 1 \tab - \tab 10 \tab 35 \tab 14 \tab - \tab 21 \tab - \tab 132 \tab 1 \tab 3 \tab 3 \tab 6 \tab 1 \tab 9 \tab - \tab - \tab 1 \tab 1 \tab 10 \tab 2 \tab 3 \tab 10 \tab - \tab 1 \tab 1 \tab 1 \tab 23 \tab 57 \tab - \tab 36 \tab 9 \tab 1 \tab 8 \tab 30 \tab 4 \tab - \tab 1 \tab - \tab 5 \tab 7 \tab 5 \tab - \tab 10 \tab 17 \tab - \tab - \tab 3 \tab 15 \tab 24 \tab 42 \tab 3 \tab 7 \tab 46 \tab 5 \tab 1 \tab 12 \tab 16 \tab 38 \tab 22 \tab 1 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab - \tab 6 \tab 4 \tab 12 \tab 3 \tab 6 \tab 27 \tab 5 \tab 2 \tab 2 \tab 17 \tab 5 \tab 2 \tab 7 \tab 2 \tab 16 \tab 24 \tab 14 \tab 14 \tab 4 \tab 5 \tab 2 \tab 2 \tab 5 \tab - \tab 14 \tab 21 \tab 20 \tab - \tab 26 \tab 524 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 3 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 7 \tab - \tab - \tab 15 \tab - \tab 6 \tab - \tab - \tab 1 \tab 1 \tab 8 \tab 3 \tab - \tab 3 \tab - \tab 11 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 8 \tab - \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 6 \tab 3 \tab 1 \tab 2 \tab 2 \tab 2 \tab 7 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 4 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 6 \tab - \tab 9 \tab 9 \tab 3 \tab 1 \tab 2 \tab 6 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab 1 \tab 1 \tab 1 \tab - \tab 5 \tab 8 \tab 6 \tab 78 \tab 2 \tab 1 \tab 4 \tab 4 \tab 1 \tab - \tab 7 \tab 5 \tab 2 \tab - \tab 4 \tab 3 \tab 20 \tab 4 \tab 2 \tab 70 \tab 2 \tab 10 \tab 12 \tab 3 \tab 20 \tab 29 \tab 3 \tab 8 \tab 63 \tab 2 \tab 11 \tab 1 \tab 2 \tab 3 \tab 32 \tab 59 \tab 4 \tab 13 \tab 93 \tab 2 \tab 39 \tab 2 \tab 4 \tab 1 \tab 1 \tab 24 \tab 3 \tab 4 \tab 2 \tab 8 \tab 50 \tab 7 \tab 9 \tab 3 \tab 8 \tab 2 \tab - \tab 6 \tab 32 \tab 2 \tab 7 \tab - \tab 17 \tab 2 \tab - \tab 9 \tab 11 \tab 4 \tab 6 \tab 9 \tab - \tab 6 \tab 1 \tab 10 \tab - \tab 3 \tab 5 \tab 6 \tab 36 \tab 14 \tab - \tab 31 \tab 2 \tab 120 \tab 4 \tab 6 \tab 11 \tab 9 \tab 5 \tab 8 \tab - \tab 4 \tab 2 \tab 5 \tab 11 \tab 5 \tab 4 \tab 8 \tab 1 \tab 4 \tab 1 \tab 4 \tab 20 \tab 51 \tab - \tab 33 \tab 6 \tab 1 \tab 3 \tab 25 \tab 9 \tab 1 \tab - \tab 2 \tab 3 \tab 13 \tab 5 \tab - \tab 10 \tab 18 \tab 1 \tab 1 \tab 3 \tab 17 \tab 12 \tab 24 \tab 5 \tab 7 \tab 20 \tab 4 \tab 1 \tab 11 \tab 22 \tab 62 \tab 28 \tab 2 \tab 6 \tab - \tab 1 \tab 6 \tab 1 \tab 1 \tab 7 \tab 5 \tab 17 \tab 2 \tab 4 \tab 20 \tab 18 \tab 5 \tab 2 \tab 12 \tab 10 \tab 1 \tab 14 \tab - \tab 27 \tab 29 \tab 12 \tab 15 \tab 10 \tab 7 \tab 1 \tab 2 \tab 2 \tab - \tab 22 \tab 23 \tab 21 \tab 3 \tab 26 \tab 625 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 4 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 7 \tab 2 \tab - \tab 11 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 8 \tab - \tab 3 \tab 7 \tab 3 \tab 12 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 5 \tab 4 \tab 2 \tab - \tab - \tab 6 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 3 \tab - \tab - \tab - \tab 9 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 11 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 5 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 10 \tab 1 \tab 11 \tab 9 \tab 2 \tab - \tab - \tab 3 \tab - \tab 4 \tab 1 \tab 2 \tab 3 \tab 4 \tab 1 \tab 4 \tab - \tab - \tab - \tab 6 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2 \tab 2 \tab 2 \tab - \tab 3 \tab 8 \tab 5 \tab 50 \tab - \tab - \tab 5 \tab 4 \tab - \tab 3 \tab 6 \tab 4 \tab 2 \tab 3 \tab 5 \tab 6 \tab 15 \tab 2 \tab 3 \tab 37 \tab 2 \tab 6 \tab 14 \tab 1 \tab 7 \tab 27 \tab 3 \tab 6 \tab 41 \tab 4 \tab 9 \tab 2 \tab 1 \tab 2 \tab 31 \tab 37 \tab 2 \tab 10 \tab 85 \tab 6 \tab 34 \tab - \tab 5 \tab 6 \tab 1 \tab 11 \tab 6 \tab 2 \tab 4 \tab 3 \tab 35 \tab 3 \tab 5 \tab 1 \tab 3 \tab 5 \tab 2 \tab 4 \tab 22 \tab 6 \tab - \tab 1 \tab 48 \tab 2 \tab - \tab 7 \tab 6 \tab 3 \tab 9 \tab 9 \tab 2 \tab 7 \tab - \tab 8 \tab - \tab 3 \tab 1 \tab 9 \tab 26 \tab 14 \tab - \tab 16 \tab 1 \tab 99 \tab 1 \tab 9 \tab 4 \tab 8 \tab 8 \tab 10 \tab - \tab 2 \tab 2 \tab 3 \tab 3 \tab 3 \tab 2 \tab 12 \tab 1 \tab 4 \tab 1 \tab 3 \tab 6 \tab 31 \tab - \tab 30 \tab 4 \tab - \tab 7 \tab 21 \tab 5 \tab 1 \tab 1 \tab - \tab 3 \tab 15 \tab 5 \tab - \tab 19 \tab 7 \tab 1 \tab 1 \tab 1 \tab 12 \tab 7 \tab 25 \tab 38 \tab 15 \tab 21 \tab - \tab 3 \tab 17 \tab 16 \tab 29 \tab 19 \tab 2 \tab 10 \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab 8 \tab 7 \tab 9 \tab - \tab 5 \tab 29 \tab 10 \tab 2 \tab 3 \tab 11 \tab 5 \tab 3 \tab 13 \tab 3 \tab 6 \tab 24 \tab 8 \tab 10 \tab 5 \tab 2 \tab 3 \tab - \tab 3 \tab 1 \tab 14 \tab 15 \tab 12 \tab 2 \tab 28 \tab 498 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab 7 \tab - \tab - \tab 3 \tab - \tab 2 \tab 2 \tab 5 \tab 3 \tab - \tab 11 \tab 2 \tab 1 \tab - \tab - \tab 4 \tab - \tab 11 \tab 2 \tab 1 \tab 5 \tab 1 \tab 7 \tab 3 \tab - \tab 2 \tab - \tab - \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 5 \tab 2 \tab - \tab - \tab 12 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 5 \tab 6 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 8 \tab - \tab - \tab - \tab 7 \tab 4 \tab 5 \tab 2 \tab 8 \tab 5 \tab 3 \tab 1 \tab - \tab 5 \tab - \tab 3 \tab 1 \tab - \tab 3 \tab 8 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 125 \tab 31 \tab 10 \tab 29 \tab 69 \tab 276 \tab 87 \tab 1574 \tab 47 \tab 44 \tab 325 \tab 194 \tab 23 \tab 22 \tab 265 \tab 59 \tab 60 \tab 69 \tab 79 \tab 81 \tab 474 \tab 95 \tab 48 \tab 905 \tab 54 \tab 93 \tab 260 \tab 176 \tab 338 \tab 595 \tab 22 \tab 155 \tab 815 \tab 161 \tab 648 \tab 26 \tab 88 \tab 156 \tab 804 \tab 1049 \tab 82 \tab 359 \tab 2029 \tab 89 \tab 1308 \tab 53 \tab 101 \tab 79 \tab 57 \tab 471 \tab 124 \tab 59 \tab 126 \tab 135 \tab 900 \tab 153 \tab 163 \tab 84 \tab 178 \tab 87 \tab 82 \tab 162 \tab 649 \tab 94 \tab 114 \tab 16 \tab 755 \tab 55 \tab 11 \tab 147 \tab 169 \tab 125 \tab 291 \tab 165 \tab 27 \tab 184 \tab 20 \tab 157 \tab 11 \tab 49 \tab 49 \tab 219 \tab 1110 \tab 324 \tab 6 \tab 545 \tab 19 \tab 6426 \tab 114 \tab 240 \tab 179 \tab 143 \tab 72 \tab 176 \tab 11 \tab 44 \tab 39 \tab 148 \tab 182 \tab 64 \tab 42 \tab 281 \tab 7 \tab 124 \tab 15 \tab 105 \tab 407 \tab 1415 \tab 15 \tab 682 \tab 216 \tab 24 \tab 99 \tab 494 \tab 133 \tab 11 \tab 7 \tab 38 \tab 60 \tab 136 \tab 78 \tab 4 \tab 259 \tab 407 \tab 19 \tab 7 \tab 48 \tab 155 \tab 299 \tab 394 \tab 125 \tab 209 \tab 728 \tab 54 \tab 38 \tab 360 \tab 240 \tab 1299 \tab 392 \tab 34 \tab 202 \tab 7 \tab 5 \tab 92 \tab 59 \tab 6 \tab 93 \tab 112 \tab 254 \tab 40 \tab 62 \tab 464 \tab 216 \tab 68 \tab 33 \tab 385 \tab 132 \tab 35 \tab 312 \tab 32 \tab 292 \tab 473 \tab 322 \tab 268 \tab 137 \tab 99 \tab 23 \tab 25 \tab 81 \tab 18 \tab 303 \tab 429 \tab 436 \tab 32 \tab 499 \tab 17254 \tab 45 \tab 57 \tab 18 \tab 20 \tab 43 \tab 39 \tab 18 \tab 27 \tab 32 \tab 101 \tab 17 \tab 17 \tab 32 \tab 75 \tab 47 \tab 3 \tab 255 \tab 33 \tab 102 \tab 36 \tab 13 \tab 46 \tab 18 \tab 176 \tab 42 \tab 14 \tab 95 \tab 33 \tab 242 \tab 37 \tab 5 \tab 29 \tab 10 \tab 26 \tab 65 \tab 31 \tab 66 \tab 40 \tab 20 \tab 29 \tab 14 \tab 11 \tab 84 \tab 47 \tab 37 \tab 39 \tab 29 \tab 29 \tab 88 \tab 22 \tab 1 \tab 3 \tab 225 \tab 32 \tab 17 \tab 12 \tab 40 \tab 24 \tab 190 \tab 41 \tab 22 \tab 4 \tab 25 \tab 19 \tab 24 \tab 15 \tab 39 \tab 11 \tab 19 \tab 16 \tab 30 \tab 8 \tab 39 \tab 10 \tab 9 \tab 46 \tab 14 \tab 13 \tab 4 \tab 38 \tab 13 \tab 30 \tab 48 \tab 8 \tab 20 \tab 47 \tab 41 \tab 42 \tab 65 \tab 14 \tab 103 \tab 116 \tab 37 \tab 16 \tab 10 \tab 58 \tab 7 \tab 48 \tab 11 \tab 23 \tab 46 \tab 65 \tab 18 \tab 20 \tab 6 \tab 7 \tab 9 \tab 52 \tab 7 \tab 7 \tab 5 \tab 8 \tab 3 \tab 17 \tab 5 \tab 30 \tab 9 \tab 7 \tab 8 \tab 10 \tab 36 \tab 5 \tab 21 \tab 9 \tab 5 \tab 1 \tab 13 \tab 12 \tab 23 \tab 14 \tab 11 \tab 44 \tab 3 \tab 1 \tab 5 \tab 10 \tab 13 \tab 15 \tab 14 \tab 14 \tab 20 \tab 19 \tab 11 \tab 4 \tab 4 \tab 10 \tab 6 \tab 3 \tab 5 \tab 9 \tab 23 \tab 10 \tab 4 \tab 5 \tab 2 \tab 4 \tab 4 \tab 3 \tab 19 \tab 22 \tab 4 \tab 6 \tab 3 \tab 18 \tab 3 \tab 7 \tab 3 \tab 4 \tab 2 \tab 4 \tab 4 \tab 3 \tab 7 \tab 8 \tab 3 \tab 1 \tab 5 \tab 6 \tab 3 \tab 1 \tab 2 \tab 14 \tab 2 \tab 1 \tab 2 \tab 1 \tab 7 \tab 3 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} i don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
