% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_07.R
\name{pawrkslf}
\alias{pawrkslf}
\title{Father self-emp. or worked for somebody}
\source{
General Social Survey https://gss.norc.org
}
\description{
pawrkslf
}
\details{
Question 8e. Was he self-employed or did he work for someone else?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab don't know \tab iap \tab no answer \tab self-employed \tab someone else \tab Total \cr
   1972 \tab 10 \tab 235 \tab 4 \tab 505 \tab 859 \tab 1613 \cr
   1973 \tab 2 \tab 152 \tab 6 \tab 494 \tab 850 \tab 1504 \cr
   1974 \tab 5 \tab 182 \tab 6 \tab 436 \tab 855 \tab 1484 \cr
   1975 \tab 3 \tab 151 \tab 21 \tab 452 \tab 863 \tab 1490 \cr
   1976 \tab - \tab 187 \tab 16 \tab 424 \tab 872 \tab 1499 \cr
   1977 \tab 2 \tab 202 \tab 4 \tab 463 \tab 859 \tab 1530 \cr
   1978 \tab - \tab 212 \tab 2 \tab 419 \tab 899 \tab 1532 \cr
   1980 \tab 1 \tab 230 \tab 11 \tab 408 \tab 818 \tab 1468 \cr
   1982 \tab 10 \tab 298 \tab 17 \tab 476 \tab 1059 \tab 1860 \cr
   1983 \tab 5 \tab 207 \tab 12 \tab 423 \tab 952 \tab 1599 \cr
   1984 \tab 5 \tab 214 \tab 3 \tab 420 \tab 831 \tab 1473 \cr
   1985 \tab 3 \tab 195 \tab 1 \tab 445 \tab 890 \tab 1534 \cr
   1986 \tab 2 \tab 224 \tab 10 \tab 383 \tab 851 \tab 1470 \cr
   1987 \tab 2 \tab 292 \tab 18 \tab 439 \tab 1068 \tab 1819 \cr
   1988 \tab 3 \tab 207 \tab 10 \tab 399 \tab 862 \tab 1481 \cr
   1989 \tab 3 \tab 225 \tab 32 \tab 393 \tab 884 \tab 1537 \cr
   1990 \tab 5 \tab 217 \tab 23 \tab 332 \tab 795 \tab 1372 \cr
   1991 \tab 9 \tab 220 \tab 47 \tab 364 \tab 877 \tab 1517 \cr
   1993 \tab 5 \tab 207 \tab 37 \tab 412 \tab 945 \tab 1606 \cr
   1994 \tab 4 \tab 508 \tab 31 \tab 695 \tab 1754 \tab 2992 \cr
   1996 \tab 14 \tab 502 \tab 40 \tab 643 \tab 1705 \tab 2904 \cr
   1998 \tab 10 \tab 503 \tab 35 \tab 630 \tab 1654 \tab 2832 \cr
   2000 \tab 8 \tab 518 \tab 35 \tab 611 \tab 1645 \tab 2817 \cr
   2002 \tab 13 \tab 526 \tab 22 \tab 586 \tab 1618 \tab 2765 \cr
   2004 \tab 15 \tab 514 \tab 24 \tab 561 \tab 1698 \tab 2812 \cr
   2006 \tab 11 \tab 2049 \tab 20 \tab 641 \tab 1789 \tab 4510 \cr
   2008 \tab 20 \tab 354 \tab 5 \tab 397 \tab 1247 \tab 2023 \cr
   2010 \tab 17 \tab 354 \tab 9 \tab 408 \tab 1256 \tab 2044 \cr
   2012 \tab 15 \tab 333 \tab 8 \tab 413 \tab 1205 \tab 1974 \cr
   2014 \tab 26 \tab 459 \tab 9 \tab 532 \tab 1512 \tab 2538 \cr
   2016 \tab 18 \tab 548 \tab 13 \tab 538 \tab 1750 \tab 2867 \cr
   2018 \tab 19 \tab 464 \tab 4 \tab 443 \tab 1418 \tab 2348 \cr
   2021 \tab - \tab 675 \tab 8 \tab 810 \tab 2539 \tab 4032 \cr
   2022 \tab - \tab 727 \tab 3 \tab 617 \tab 2197 \tab 3544 \cr
   Total \tab 265 \tab 13091 \tab 546 \tab 16612 \tab 41876 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} self-employed
\item \code{2} someone else
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
