% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_78.R
\name{phone}
\alias{phone}
\title{Does r have telephone}
\source{
General Social Survey https://gss.norc.org
}
\description{
phone
}
\details{
Question 243. Thank you very much for your time and help. Do you have a telephone?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab home phone \tab no answer \tab no phone number \tab phone elsewhere \tab refused phone number \tab phone location not given \tab cell phone \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 1302 \tab 12 \tab 115 \tab 27 \tab 48 \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 1277 \tab 6 \tab 117 \tab 26 \tab 48 \tab 10 \tab - \tab 1484 \cr
   1975 \tab - \tab 1243 \tab 2 \tab 103 \tab 43 \tab 41 \tab 58 \tab - \tab 1490 \cr
   1976 \tab - \tab 1272 \tab - \tab 153 \tab 9 \tab 47 \tab 18 \tab - \tab 1499 \cr
   1977 \tab - \tab 1324 \tab 5 \tab 124 \tab 47 \tab 25 \tab 5 \tab - \tab 1530 \cr
   1978 \tab - \tab 1359 \tab 1 \tab 87 \tab 40 \tab 31 \tab 14 \tab - \tab 1532 \cr
   1980 \tab - \tab 1316 \tab 2 \tab 92 \tab 32 \tab 18 \tab 8 \tab - \tab 1468 \cr
   1982 \tab - \tab 1609 \tab 6 \tab 131 \tab 38 \tab 44 \tab 32 \tab - \tab 1860 \cr
   1983 \tab - \tab 1380 \tab 6 \tab 94 \tab 58 \tab 57 \tab 4 \tab - \tab 1599 \cr
   1984 \tab - \tab 1267 \tab 5 \tab 91 \tab 66 \tab 44 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 1374 \tab 1 \tab 71 \tab 38 \tab 50 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 1320 \tab 2 \tab 84 \tab 44 \tab 20 \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 1606 \tab 2 \tab 111 \tab 60 \tab 40 \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 1336 \tab 2 \tab 70 \tab 42 \tab 31 \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 1376 \tab 8 \tab 72 \tab 43 \tab 38 \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 1239 \tab 9 \tab 63 \tab 38 \tab 23 \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 1345 \tab 19 \tab 79 \tab 42 \tab 32 \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 1463 \tab 5 \tab 55 \tab 54 \tab 29 \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 2757 \tab 7 \tab 111 \tab 58 \tab 58 \tab 1 \tab - \tab 2992 \cr
   1996 \tab - \tab 2673 \tab - \tab 89 \tab 76 \tab 65 \tab 1 \tab - \tab 2904 \cr
   1998 \tab - \tab 2516 \tab 6 \tab 114 \tab 77 \tab 114 \tab 5 \tab - \tab 2832 \cr
   2000 \tab - \tab 2532 \tab 3 \tab 91 \tab 93 \tab 94 \tab 4 \tab - \tab 2817 \cr
   2002 \tab - \tab 2498 \tab - \tab 17 \tab 116 \tab 132 \tab 2 \tab - \tab 2765 \cr
   2004 \tab - \tab 2418 \tab 34 \tab 39 \tab 70 \tab 38 \tab 5 \tab 208 \tab 2812 \cr
   2006 \tab - \tab 3753 \tab 6 \tab 25 \tab 85 \tab 190 \tab - \tab 451 \tab 4510 \cr
   2008 \tab - \tab 1556 \tab 11 \tab 52 \tab 26 \tab 157 \tab - \tab 221 \tab 2023 \cr
   2010 \tab - \tab 1172 \tab 2 \tab 19 \tab 93 \tab 71 \tab - \tab 687 \tab 2044 \cr
   2012 \tab - \tab 984 \tab - \tab 11 \tab 78 \tab 69 \tab 1 \tab 831 \tab 1974 \cr
   2014 \tab - \tab 1075 \tab 5 \tab 12 \tab 68 \tab 66 \tab - \tab 1312 \tab 2538 \cr
   2016 \tab - \tab 1010 \tab - \tab 19 \tab 86 \tab 80 \tab - \tab 1672 \tab 2867 \cr
   2018 \tab - \tab 713 \tab - \tab 15 \tab 35 \tab 89 \tab - \tab 1496 \tab 2348 \cr
   2021 \tab - \tab - \tab 308 \tab 633 \tab - \tab 712 \tab - \tab 2379 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab 595 \tab - \tab 489 \tab - \tab 2460 \tab 3544 \cr
   Total \tab 1613 \tab 50065 \tab 475 \tab 3554 \tab 1708 \tab 3090 \tab 168 \tab 11717 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} no phone number
\item \code{2} refused phone number
\item \code{3} home phone
\item \code{4} phone elsewhere
\item \code{5} phone location not given
\item \code{6} cell phone
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
