% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_76.R
\name{pistol}
\alias{pistol}
\title{pistol}
\source{
General Social Survey https://gss.norc.org
}
\description{
Pistol or revolver in home
}
\details{
Question 237a. If yes: Is it a pistol, shotgun, rigfle, or what? PISTOL OR REVOLVER
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab no answer \tab refused \tab yes \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 1184 \tab 7 \tab 15 \tab 298 \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 1176 \tab 5 \tab 10 \tab 293 \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 1151 \tab 10 \tab 17 \tab 321 \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 1204 \tab 11 \tab 2 \tab 313 \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 1113 \tab 10 \tab 7 \tab 338 \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 1434 \tab 6 \tab 26 \tab 394 \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 1144 \tab 6 \tab 9 \tab 314 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 1170 \tab 4 \tab 10 \tab 350 \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 1371 \tab 7 \tab 10 \tab 431 \tab - \tab - \tab 1819 \cr
   1988 \tab 504 \tab 741 \tab 7 \tab 10 \tab 219 \tab - \tab - \tab 1481 \cr
   1989 \tab 504 \tab 768 \tab 4 \tab - \tab 261 \tab - \tab - \tab 1537 \cr
   1990 \tab 455 \tab 694 \tab 9 \tab - \tab 213 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 774 \tab 20 \tab - \tab 199 \tab - \tab - \tab 1517 \cr
   1993 \tab 531 \tab 809 \tab 2 \tab 7 \tab 257 \tab - \tab - \tab 1606 \cr
   1994 \tab 996 \tab 1485 \tab 23 \tab 5 \tab 483 \tab - \tab - \tab 2992 \cr
   1996 \tab 981 \tab 1484 \tab 5 \tab 7 \tab 427 \tab - \tab - \tab 2904 \cr
   1998 \tab 951 \tab 1500 \tab 9 \tab 6 \tab 366 \tab - \tab - \tab 2832 \cr
   2000 \tab 956 \tab 1459 \tab 14 \tab 23 \tab 364 \tab 1 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 734 \tab - \tab 10 \tab 180 \tab - \tab - \tab 2765 \cr
   2004 \tab 1916 \tab 686 \tab 11 \tab 16 \tab 183 \tab - \tab - \tab 2812 \cr
   2006 \tab 2509 \tab 1563 \tab - \tab 35 \tab 400 \tab 3 \tab - \tab 4510 \cr
   2008 \tab 674 \tab 1040 \tab - \tab 19 \tab 286 \tab 4 \tab - \tab 2023 \cr
   2010 \tab 766 \tab 984 \tab - \tab 50 \tab 239 \tab 5 \tab - \tab 2044 \cr
   2012 \tab 669 \tab 981 \tab - \tab 28 \tab 289 \tab 7 \tab - \tab 1974 \cr
   2014 \tab 827 \tab 1277 \tab - \tab 63 \tab 366 \tab 5 \tab - \tab 2538 \cr
   2016 \tab 978 \tab 1417 \tab - \tab 60 \tab 406 \tab 6 \tab - \tab 2867 \cr
   2018 \tab 779 \tab 1117 \tab 4 \tab 39 \tab 405 \tab 4 \tab - \tab 2348 \cr
   2021 \tab - \tab 2818 \tab 4 \tab 117 \tab 1008 \tab 76 \tab 9 \tab 4032 \cr
   2022 \tab 1203 \tab 1671 \tab 6 \tab 72 \tab 561 \tab 24 \tab 7 \tab 3544 \cr
   Total \tab 26268 \tab 34949 \tab 184 \tab 673 \tab 10164 \tab 136 \tab 16 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} refused
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
