% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{polviews}
\alias{polviews}
\title{Think of self as liberal or conservative}
\source{
General Social Survey https://gss.norc.org
}
\description{
polviews
}
\details{
Question 67a. We hear a lot of talk these days about liberals and conservatives. I'm going to show you a seven-point scale on which the political views that people might hold are arranged from extremely liberal - point 1 - to extremely conservative - point 7. Where would you place yourself on this scale?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab conservative \tab don't know \tab extremely conservative \tab extremely liberal \tab liberal \tab moderate, middle of the road \tab no answer \tab slightly conservative \tab slightly liberal \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 160 \tab 70 \tab 35 \tab 22 \tab 201 \tab 564 \tab 4 \tab 221 \tab 207 \tab - \tab 1484 \cr
   1975 \tab - \tab 150 \tab 81 \tab 35 \tab 46 \tab 179 \tab 559 \tab 12 \tab 232 \tab 196 \tab - \tab 1490 \cr
   1976 \tab - \tab 190 \tab 93 \tab 27 \tab 31 \tab 187 \tab 559 \tab 5 \tab 221 \tab 186 \tab - \tab 1499 \cr
   1977 \tab - \tab 179 \tab 71 \tab 39 \tab 37 \tab 169 \tab 564 \tab 6 \tab 251 \tab 214 \tab - \tab 1530 \cr
   1978 \tab - \tab 188 \tab 70 \tab 30 \tab 22 \tab 142 \tab 549 \tab 27 \tab 263 \tab 241 \tab - \tab 1532 \cr
   1980 \tab - \tab 180 \tab 22 \tab 44 \tab 36 \tab 121 \tab 582 \tab 17 \tab 258 \tab 208 \tab - \tab 1468 \cr
   1982 \tab - \tab 229 \tab 104 \tab 67 \tab 48 \tab 199 \tab 694 \tab 17 \tab 235 \tab 267 \tab - \tab 1860 \cr
   1983 \tab 795 \tab 108 \tab 31 \tab 20 \tab 16 \tab 67 \tab 319 \tab 3 \tab 142 \tab 98 \tab - \tab 1599 \cr
   1984 \tab - \tab 186 \tab 52 \tab 41 \tab 29 \tab 133 \tab 568 \tab 11 \tab 276 \tab 177 \tab - \tab 1473 \cr
   1985 \tab - \tab 214 \tab 63 \tab 42 \tab 35 \tab 163 \tab 566 \tab 9 \tab 271 \tab 171 \tab - \tab 1534 \cr
   1986 \tab - \tab 209 \tab 67 \tab 38 \tab 25 \tab 132 \tab 579 \tab 2 \tab 242 \tab 176 \tab - \tab 1470 \cr
   1987 \tab - \tab 208 \tab 98 \tab 45 \tab 47 \tab 234 \tab 641 \tab 42 \tab 278 \tab 226 \tab - \tab 1819 \cr
   1988 \tab - \tab 225 \tab 56 \tab 32 \tab 34 \tab 176 \tab 514 \tab 9 \tab 246 \tab 189 \tab - \tab 1481 \cr
   1989 \tab - \tab 194 \tab 88 \tab 27 \tab 40 \tab 178 \tab 566 \tab 7 \tab 246 \tab 191 \tab - \tab 1537 \cr
   1990 \tab - \tab 192 \tab 53 \tab 51 \tab 36 \tab 141 \tab 476 \tab 4 \tab 240 \tab 179 \tab - \tab 1372 \cr
   1991 \tab - \tab 212 \tab 54 \tab 40 \tab 37 \tab 154 \tab 584 \tab 4 \tab 218 \tab 214 \tab - \tab 1517 \cr
   1993 \tab - \tab 254 \tab 49 \tab 42 \tab 30 \tab 180 \tab 575 \tab 9 \tab 264 \tab 203 \tab - \tab 1606 \cr
   1994 \tab - \tab 478 \tab 101 \tab 103 \tab 71 \tab 328 \tab 1049 \tab 12 \tab 472 \tab 378 \tab - \tab 2992 \cr
   1996 \tab - \tab 458 \tab 155 \tab 93 \tab 59 \tab 303 \tab 1045 \tab 6 \tab 451 \tab 334 \tab - \tab 2904 \cr
   1998 \tab - \tab 415 \tab 133 \tab 86 \tab 64 \tab 357 \tab 986 \tab 8 \tab 432 \tab 351 \tab - \tab 2832 \cr
   2000 \tab - \tab 411 \tab 149 \tab 89 \tab 107 \tab 308 \tab 1054 \tab 24 \tab 390 \tab 285 \tab - \tab 2817 \cr
   2002 \tab 1393 \tab 210 \tab 36 \tab 41 \tab 47 \tab 143 \tab 522 \tab 5 \tab 209 \tab 159 \tab - \tab 2765 \cr
   2004 \tab 1472 \tab 223 \tab 26 \tab 56 \tab 46 \tab 120 \tab 497 \tab 5 \tab 214 \tab 153 \tab - \tab 2812 \cr
   2006 \tab - \tab 685 \tab 154 \tab 167 \tab 139 \tab 524 \tab 1683 \tab 23 \tab 618 \tab 517 \tab - \tab 4510 \cr
   2008 \tab - \tab 327 \tab 77 \tab 68 \tab 69 \tab 240 \tab 740 \tab 13 \tab 268 \tab 221 \tab - \tab 2023 \cr
   2010 \tab - \tab 315 \tab 61 \tab 80 \tab 76 \tab 259 \tab 746 \tab 10 \tab 265 \tab 232 \tab - \tab 2044 \cr
   2012 \tab - \tab 292 \tab 87 \tab 68 \tab 81 \tab 244 \tab 713 \tab 13 \tab 268 \tab 208 \tab - \tab 1974 \cr
   2014 \tab - \tab 358 \tab 65 \tab 107 \tab 94 \tab 304 \tab 989 \tab 24 \tab 334 \tab 263 \tab - \tab 2538 \cr
   2016 \tab - \tab 426 \tab 81 \tab 120 \tab 136 \tab 350 \tab 1032 \tab 30 \tab 382 \tab 310 \tab - \tab 2867 \cr
   2018 \tab - \tab 354 \tab 79 \tab 99 \tab 122 \tab 278 \tab 855 \tab 22 \tab 283 \tab 256 \tab - \tab 2348 \cr
   2021 \tab 15 \tab 617 \tab 10 \tab 174 \tab 207 \tab 623 \tab 1377 \tab 3 \tab 476 \tab 490 \tab 40 \tab 4032 \cr
   2022 \tab 1 \tab 514 \tab 73 \tab 159 \tab 192 \tab 486 \tab 1245 \tab 27 \tab 430 \tab 400 \tab 17 \tab 3544 \cr
   Total \tab 6793 \tab 9361 \tab 2409 \tab 2165 \tab 2081 \tab 7623 \tab 23992 \tab 413 \tab 9596 \tab 7900 \tab 57 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} extremely liberal
\item \code{2} liberal
\item \code{3} slightly liberal
\item \code{4} moderate, middle of the road
\item \code{5} slightly conservative
\item \code{6} conservative
\item \code{7} extremely conservative
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
