% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_39.R
\name{popespks}
\alias{popespks}
\title{Pope is infallible on matter's of faith or morals}
\source{
General Social Survey https://gss.norc.org
}
\description{
popespks
}
\details{
Question 114g. We are interested in what American Catholics think about religious matters. Please select the answer from this card that comes closest to your own personal opinion about the following statement: under certain conditions, the pope is infallible when he speaks on matters of faith and morals.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab certainly false \tab certainly true \tab don't know \tab i am uncertain whether this is true or false \tab probably false \tab probably true \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2492 \tab 12 \tab 82 \tab 5 \tab 111 \tab 22 \tab 88 \tab - \tab - \tab 2812 \cr
   2006 \tab 3756 \tab 40 \tab 187 \tab 7 \tab 243 \tab 68 \tab 191 \tab 18 \tab - \tab 4510 \cr
   2008 \tab 1790 \tab 9 \tab 45 \tab 3 \tab 88 \tab 16 \tab 65 \tab 7 \tab - \tab 2023 \cr
   2010 \tab 1561 \tab 23 \tab 84 \tab 12 \tab 167 \tab 45 \tab 149 \tab 3 \tab - \tab 2044 \cr
   2012 \tab 1529 \tab 19 \tab 89 \tab 15 \tab 170 \tab 41 \tab 105 \tab 6 \tab - \tab 1974 \cr
   2014 \tab 1918 \tab 26 \tab 133 \tab 9 \tab 222 \tab 36 \tab 179 \tab 15 \tab - \tab 2538 \cr
   2016 \tab 2204 \tab 13 \tab 155 \tab 8 \tab 210 \tab 53 \tab 208 \tab 16 \tab - \tab 2867 \cr
   2018 \tab 1839 \tab 17 \tab 107 \tab 7 \tab 167 \tab 34 \tab 159 \tab 18 \tab - \tab 2348 \cr
   2021 \tab 3203 \tab 55 \tab 132 \tab 5 \tab 327 \tab 84 \tab 215 \tab 7 \tab 4 \tab 4032 \cr
   2022 \tab 2771 \tab 57 \tab 120 \tab 12 \tab 320 \tab 68 \tab 160 \tab 31 \tab 5 \tab 3544 \cr
   Total \tab 66761 \tab 271 \tab 1134 \tab 83 \tab 2025 \tab 467 \tab 1519 \tab 121 \tab 9 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} certainly true
\item \code{2} probably true
\item \code{3} i am uncertain whether this is true or false
\item \code{4} probably false
\item \code{5} certainly false
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
