% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_57.R
\name{popular}
\alias{popular}
\title{popular}
\source{
General Social Survey https://gss.norc.org
}
\description{
To be well liked or popular
}
\details{
Question 170. If you had to choose, which thing on this list would you pick as the most important for a child to learn to prepare him or her for life? a. Which comes next in importantc? b. Which comes third? c. Which comes fourth? TO BE WELL-LIKED OR POPULAR
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab fifth \tab first \tab fourth \tab no answer \tab second \tab third \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 730 \tab 525 \tab 4 \tab 142 \tab 8 \tab 28 \tab 33 \tab 1470 \cr
   1987 \tab - \tab 1217 \tab 15 \tab 392 \tab 19 \tab 77 \tab 99 \tab 1819 \cr
   1988 \tab 484 \tab 683 \tab 7 \tab 188 \tab 20 \tab 45 \tab 54 \tab 1481 \cr
   1989 \tab 502 \tab 713 \tab 11 \tab 188 \tab 35 \tab 33 \tab 55 \tab 1537 \cr
   1990 \tab 473 \tab 614 \tab 8 \tab 181 \tab 28 \tab 18 \tab 50 \tab 1372 \cr
   1991 \tab 500 \tab 709 \tab 4 \tab 185 \tab 35 \tab 27 \tab 57 \tab 1517 \cr
   1993 \tab 549 \tab 752 \tab 11 \tab 193 \tab 26 \tab 26 \tab 49 \tab 1606 \cr
   1994 \tab 981 \tab 1518 \tab 9 \tab 330 \tab 33 \tab 43 \tab 78 \tab 2992 \cr
   1996 \tab 979 \tab 1439 \tab 13 \tab 344 \tab 30 \tab 35 \tab 64 \tab 2904 \cr
   1998 \tab 921 \tab 1424 \tab 19 \tab 330 \tab 31 \tab 41 \tab 66 \tab 2832 \cr
   2000 \tab 921 \tab 1422 \tab 9 \tab 335 \tab 33 \tab 26 \tab 71 \tab 2817 \cr
   2002 \tab 1853 \tab 674 \tab 4 \tab 178 \tab 13 \tab 11 \tab 32 \tab 2765 \cr
   2004 \tab 1936 \tab 665 \tab 7 \tab 153 \tab 8 \tab 14 \tab 29 \tab 2812 \cr
   2006 \tab 3508 \tab 792 \tab 7 \tab 155 \tab 7 \tab 13 \tab 28 \tab 4510 \cr
   2008 \tab 658 \tab 1033 \tab 12 \tab 254 \tab 14 \tab 14 \tab 38 \tab 2023 \cr
   2010 \tab 667 \tab 1033 \tab 5 \tab 243 \tab 28 \tab 22 \tab 46 \tab 2044 \cr
   2012 \tab 636 \tab 1008 \tab 7 \tab 255 \tab 15 \tab 21 \tab 32 \tab 1974 \cr
   2014 \tab 851 \tab 1266 \tab 4 \tab 329 \tab 17 \tab 20 \tab 51 \tab 2538 \cr
   2016 \tab 911 \tab 1473 \tab 11 \tab 376 \tab 22 \tab 23 \tab 51 \tab 2867 \cr
   2018 \tab 785 \tab 1172 \tab 5 \tab 306 \tab 21 \tab 17 \tab 42 \tab 2348 \cr
   2021 \tab 1360 \tab 1821 \tab 21 \tab 663 \tab 99 \tab 20 \tab 48 \tab 4032 \cr
   2022 \tab 1173 \tab 1592 \tab 18 \tab 593 \tab 77 \tab 24 \tab 67 \tab 3544 \cr
   Total \tab 39964 \tab 23545 \tab 211 \tab 6313 \tab 619 \tab 598 \tab 1140 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} first
\item \code{2} second
\item \code{3} third
\item \code{4} fourth
\item \code{5} fifth
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
