% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_71.R
\name{pornlaw}
\alias{pornlaw}
\title{Feelings about pornography laws}
\source{
General Social Survey https://gss.norc.org
}
\description{
pornlaw
}
\details{
Question 222. Which of these statements comes closest to your feelings about pornography laws: 1. There should be laws against the distribution of pornography whatever the age. 2. There should be laws against the distribution of pornography to persons under 18. 3. There should be no laws forbidding the distribution of pornography.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab don't know \tab no answer \tab there should be laws against the distribution of pornography to persons under 18 \tab there should be laws against the distribution of pornography whatever the age \tab there should be no laws forbidding the distribution of pornography \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 29 \tab 6 \tab 705 \tab 626 \tab 138 \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 18 \tab 1 \tab 709 \tab 602 \tab 160 \tab - \tab 1490 \cr
   1976 \tab - \tab 31 \tab 3 \tab 745 \tab 599 \tab 121 \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 16 \tab 1 \tab 743 \tab 661 \tab 111 \tab - \tab 1532 \cr
   1980 \tab - \tab 32 \tab - \tab 755 \tab 591 \tab 90 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 21 \tab 4 \tab 847 \tab 656 \tab 71 \tab - \tab 1599 \cr
   1984 \tab - \tab 18 \tab 9 \tab 790 \tab 595 \tab 61 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 14 \tab 1 \tab 776 \tab 626 \tab 53 \tab - \tab 1470 \cr
   1987 \tab - \tab 29 \tab 8 \tab 1023 \tab 681 \tab 78 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 19 \tab 3 \tab 499 \tab 429 \tab 47 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 14 \tab 3 \tab 552 \tab 416 \tab 50 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 13 \tab - \tab 471 \tab 365 \tab 50 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 15 \tab 3 \tab 558 \tab 401 \tab 40 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 14 \tab 3 \tab 568 \tab 438 \tab 34 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 15 \tab 12 \tab 1189 \tab 736 \tab 59 \tab - \tab 2992 \cr
   1996 \tab 979 \tab 21 \tab 3 \tab 1109 \tab 720 \tab 72 \tab - \tab 2904 \cr
   1998 \tab 921 \tab 19 \tab 5 \tab 1094 \tab 720 \tab 73 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 21 \tab 10 \tab 1136 \tab 670 \tab 59 \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 6 \tab 4 \tab 512 \tab 343 \tab 47 \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 4 \tab 3 \tab 493 \tab 339 \tab 37 \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 19 \tab 3 \tab 1125 \tab 775 \tab 67 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 6 \tab - \tab 848 \tab 460 \tab 51 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 9 \tab 3 \tab 874 \tab 436 \tab 55 \tab - \tab 2044 \cr
   2012 \tab 636 \tab 12 \tab 3 \tab 825 \tab 455 \tab 43 \tab - \tab 1974 \cr
   2014 \tab 851 \tab 8 \tab 3 \tab 1042 \tab 563 \tab 71 \tab - \tab 2538 \cr
   2016 \tab 911 \tab 12 \tab 6 \tab 1244 \tab 611 \tab 83 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 10 \tab 3 \tab 978 \tab 492 \tab 80 \tab - \tab 2348 \cr
   2021 \tab 1361 \tab 2 \tab - \tab 1828 \tab 686 \tab 138 \tab 17 \tab 4032 \cr
   2022 \tab 1173 \tab 20 \tab 13 \tab 1540 \tab 672 \tab 114 \tab 12 \tab 3544 \cr
   Total \tab 27683 \tab 467 \tab 116 \tab 25578 \tab 16364 \tab 2153 \tab 29 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} there should be laws against the distribution of pornography whatever the age
\item \code{2} there should be laws against the distribution of pornography to persons under 18
\item \code{3} there should be no laws forbidding the distribution of pornography
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
