% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_70.R
\name{pornmorl}
\alias{pornmorl}
\title{pornmorl}
\source{
General Social Survey https://gss.norc.org
}
\description{
Materials lead to breakdown in morals
}
\details{
Question 221. The next questions are about pornography - books, movies, magazines, and photographs that show or describe sex activities. I'm going to read some opinions about the effects of looking at or reading such sexual materials. As I read each one, please tell me if you think sexual materials do or do not have that effect: b. Sexual materials lead to breakdown of morals.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 84 \tab 622 \tab 4 \tab 794 \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 134 \tab 588 \tab 3 \tab 765 \tab 1490 \cr
   1976 \tab - \tab 111 \tab 560 \tab 9 \tab 819 \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 78 \tab 581 \tab 3 \tab 870 \tab 1532 \cr
   1980 \tab - \tab 103 \tab 477 \tab 7 \tab 881 \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 83 \tab 573 \tab 10 \tab 933 \tab 1599 \cr
   1984 \tab - \tab 75 \tab 484 \tab 11 \tab 903 \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 78 \tab 475 \tab 7 \tab 910 \tab 1470 \cr
   1987 \tab - \tab 124 \tab 613 \tab 9 \tab 1073 \tab 1819 \cr
   1988 \tab 484 \tab 51 \tab 325 \tab 5 \tab 616 \tab 1481 \cr
   1989 \tab 502 \tab 62 \tab 323 \tab 7 \tab 643 \tab 1537 \cr
   1990 \tab 473 \tab 64 \tab 287 \tab 1 \tab 547 \tab 1372 \cr
   1991 \tab 500 \tab 69 \tab 340 \tab 2 \tab 606 \tab 1517 \cr
   1993 \tab 549 \tab 58 \tab 318 \tab 10 \tab 671 \tab 1606 \cr
   1994 \tab 1965 \tab 71 \tab 370 \tab 4 \tab 582 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 52504 \tab 1245 \tab 6936 \tab 92 \tab 11613 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
