% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_71.R
\name{pornout}
\alias{pornout}
\title{Materials provide outlet}
\source{
General Social Survey https://gss.norc.org
}
\description{
pornout
}
\details{
Question 221. The next questions are about pornography - books, movies, magazines, and photographs that show or describe sex activities. I'm going to read some opinions about the effects of looking at or reading such sexual materials. As I read each one, please tell me if you think sexual materials do or do not have that effect: d. Sexual materials provide an outlet for bottled-up impulses
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 151 \tab 528 \tab 5 \tab 820 \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 236 \tab 413 \tab 5 \tab 836 \tab 1490 \cr
   1976 \tab - \tab 235 \tab 427 \tab 8 \tab 829 \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 183 \tab 445 \tab 5 \tab 899 \tab 1532 \cr
   1980 \tab - \tab 204 \tab 393 \tab 6 \tab 865 \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 176 \tab 501 \tab 11 \tab 911 \tab 1599 \cr
   1984 \tab - \tab 177 \tab 399 \tab 10 \tab 887 \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 145 \tab 447 \tab 8 \tab 870 \tab 1470 \cr
   1987 \tab - \tab 253 \tab 479 \tab 12 \tab 1075 \tab 1819 \cr
   1988 \tab 484 \tab 132 \tab 303 \tab 5 \tab 557 \tab 1481 \cr
   1989 \tab 502 \tab 141 \tab 271 \tab 6 \tab 617 \tab 1537 \cr
   1990 \tab 473 \tab 129 \tab 239 \tab 3 \tab 528 \tab 1372 \cr
   1991 \tab 500 \tab 139 \tab 307 \tab 4 \tab 567 \tab 1517 \cr
   1993 \tab 549 \tab 138 \tab 275 \tab 10 \tab 634 \tab 1606 \cr
   1994 \tab 1965 \tab 132 \tab 303 \tab 5 \tab 587 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 52504 \tab 2571 \tab 5730 \tab 103 \tab 11482 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
