% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_37.R
\name{pray}
\alias{pray}
\title{pray}
\source{
General Social Survey https://gss.norc.org
}
\description{
How often does r pray
}
\details{
Question 110. About how often do you pray?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab don't know \tab less than once a week \tab never \tab no answer \tab once a day \tab once a week \tab several times a day \tab several times a week \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 6 \tab 325 \tab 61 \tab 20 \tab 474 \tab 122 \tab 387 \tab 204 \tab - \tab 1599 \cr
   1984 \tab - \tab 2 \tab 292 \tab 20 \tab 24 \tab 421 \tab 104 \tab 409 \tab 201 \tab - \tab 1473 \cr
   1985 \tab - \tab - \tab 311 \tab 19 \tab 24 \tab 468 \tab 112 \tab 410 \tab 190 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab 324 \tab 9 \tab 30 \tab 587 \tab 137 \tab 484 \tab 248 \tab - \tab 1819 \cr
   1988 \tab - \tab 5 \tab 326 \tab 4 \tab 8 \tab 450 \tab 113 \tab 346 \tab 229 \tab - \tab 1481 \cr
   1989 \tab 531 \tab - \tab 248 \tab 6 \tab 12 \tab 294 \tab 66 \tab 230 \tab 150 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 2 \tab 229 \tab - \tab 25 \tab 261 \tab 77 \tab 210 \tab 124 \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 526 \tab - \tab 224 \tab 16 \tab 24 \tab 311 \tab 81 \tab 281 \tab 143 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 3 \tab 440 \tab 26 \tab 21 \tab 640 \tab 167 \tab 454 \tab 226 \tab - \tab 2992 \cr
   1996 \tab 1914 \tab 1 \tab 178 \tab 25 \tab 5 \tab 312 \tab 79 \tab 256 \tab 134 \tab - \tab 2904 \cr
   1998 \tab 1387 \tab 3 \tab 313 \tab 24 \tab 14 \tab 426 \tab 108 \tab 355 \tab 202 \tab - \tab 2832 \cr
   2000 \tab 1419 \tab 1 \tab 309 \tab 12 \tab 12 \tab 402 \tab 90 \tab 376 \tab 196 \tab - \tab 2817 \cr
   2002 \tab 1393 \tab 10 \tab 302 \tab 7 \tab 10 \tab 424 \tab 117 \tab 348 \tab 154 \tab - \tab 2765 \cr
   2004 \tab 1472 \tab 4 \tab 144 \tab 137 \tab 4 \tab 369 \tab 77 \tab 416 \tab 189 \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 11 \tab 342 \tab 316 \tab 10 \tab 863 \tab 192 \tab 926 \tab 332 \tab - \tab 4510 \cr
   2008 \tab - \tab 11 \tab 256 \tab 235 \tab 10 \tab 596 \tab 116 \tab 576 \tab 223 \tab - \tab 2023 \cr
   2010 \tab - \tab 12 \tab 222 \tab 264 \tab 9 \tab 592 \tab 118 \tab 587 \tab 240 \tab - \tab 2044 \cr
   2012 \tab - \tab 9 \tab 199 \tab 261 \tab 10 \tab 592 \tab 130 \tab 574 \tab 199 \tab - \tab 1974 \cr
   2014 \tab - \tab 7 \tab 263 \tab 374 \tab 12 \tab 763 \tab 153 \tab 708 \tab 258 \tab - \tab 2538 \cr
   2016 \tab - \tab 9 \tab 257 \tab 424 \tab 20 \tab 827 \tab 171 \tab 857 \tab 302 \tab - \tab 2867 \cr
   2018 \tab - \tab 5 \tab 251 \tab 358 \tab 10 \tab 674 \tab 144 \tab 691 \tab 215 \tab - \tab 2348 \cr
   2021 \tab - \tab 2 \tab 560 \tab 883 \tab 41 \tab 656 \tab 175 \tab 1139 \tab 542 \tab 34 \tab 4032 \cr
   2022 \tab - \tab 12 \tab 453 \tab 731 \tab 17 \tab 681 \tab 189 \tab 1035 \tab 412 \tab 14 \tab 3544 \cr
   Total \tab 28586 \tab 115 \tab 6768 \tab 4212 \tab 372 \tab 12083 \tab 2838 \tab 12055 \tab 5313 \tab 48 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} several times a day
\item \code{2} once a day
\item \code{3} several times a week
\item \code{4} once a week
\item \code{5} less than once a week
\item \code{6} never
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
