% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_42.R
\name{prayer}
\alias{prayer}
\title{prayer}
\source{
General Social Survey https://gss.norc.org
}
\description{
Bible prayer in public schools
}
\details{
Question 119a. The United States Supreme Court has ruled that no state or local government may require the reading of the Lor's Prayer or Bible verses in public schools. What are your views on this - do you approve or disapprove of the court ruling?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab approve \tab disapprove \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 731 \tab 231 \tab 496 \tab 23 \tab 3 \tab - \tab 1484 \cr
   1975 \tab - \tab 526 \tab 918 \tab 43 \tab 3 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 511 \tab 981 \tab 35 \tab 3 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 635 \tab 1161 \tab 60 \tab 4 \tab - \tab 1860 \cr
   1983 \tab - \tab 631 \tab 904 \tab 60 \tab 4 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 659 \tab 824 \tab 47 \tab 4 \tab - \tab 1534 \cr
   1986 \tab 740 \tab 270 \tab 442 \tab 11 \tab 7 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 368 \tab 581 \tab 37 \tab 2 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 411 \tab 562 \tab 32 \tab 1 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 367 \tab 517 \tab 42 \tab 2 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 392 \tab 594 \tab 35 \tab 3 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 423 \tab 611 \tab 43 \tab 3 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 737 \tab 1150 \tab 85 \tab 5 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 780 \tab 1099 \tab 79 \tab 2 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 794 \tab 988 \tab 85 \tab 4 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 688 \tab 1087 \tab 94 \tab 8 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 341 \tab 518 \tab 49 \tab - \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 319 \tab 558 \tab 28 \tab 1 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 796 \tab 1094 \tab 101 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 517 \tab 758 \tab 54 \tab - \tab - \tab 2023 \cr
   2010 \tab 614 \tab 598 \tab 767 \tab 64 \tab 1 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 510 \tab 743 \tab 46 \tab 3 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 655 \tab 949 \tab 71 \tab - \tab - \tab 2538 \cr
   2016 \tab 979 \tab 753 \tab 1057 \tab 75 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 689 \tab 782 \tab 86 \tab 2 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 2324 \tab 589 \tab 554 \tab 73 \tab 4 \tab - \tab 3544 \cr
   Total \tab 31942 \tab 14190 \tab 20695 \tab 1458 \tab 73 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} approve
\item \code{2} disapprove
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
