% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_70.R
\name{premarsx}
\alias{premarsx}
\title{Sex before marriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
premarsx
}
\details{
Question 217. There's been a lot of discussion about the way morals and attitudes about sex are changing in this country. If a man and woman have sex relations before marriage, do you think it is always wrong, almost always wrong, wrong only sometimes, or not wrong at all?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab almost always wrong \tab always wrong \tab don't know \tab no answer \tab not wrong at all \tab wrong only sometimes \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 182 \tab 563 \tab 65 \tab 11 \tab 419 \tab 373 \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \tab - \tab 1504 \cr
   1974 \tab 181 \tab 472 \tab 48 \tab 7 \tab 439 \tab 337 \tab - \tab - \tab 1484 \cr
   1975 \tab 176 \tab 441 \tab 58 \tab 5 \tab 468 \tab 342 \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \tab - \tab 1499 \cr
   1977 \tab 141 \tab 459 \tab 39 \tab 10 \tab 541 \tab 340 \tab - \tab - \tab 1530 \cr
   1978 \tab 175 \tab 437 \tab 34 \tab 4 \tab 578 \tab 304 \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \tab - \tab 1468 \cr
   1982 \tab 152 \tab 520 \tab 55 \tab 11 \tab 763 \tab 359 \tab - \tab - \tab 1860 \cr
   1983 \tab 155 \tab 427 \tab 31 \tab 7 \tab 601 \tab 378 \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \tab - \tab 1473 \cr
   1985 \tab 123 \tab 422 \tab 48 \tab 4 \tab 643 \tab 294 \tab - \tab - \tab 1534 \cr
   1986 \tab 127 \tab 396 \tab 39 \tab 6 \tab 574 \tab 328 \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \tab - \tab 1819 \cr
   1988 \tab 102 \tab 251 \tab 27 \tab 6 \tab 388 \tab 214 \tab 493 \tab - \tab 1481 \cr
   1989 \tab 84 \tab 272 \tab 31 \tab 4 \tab 394 \tab 221 \tab 531 \tab - \tab 1537 \cr
   1990 \tab 103 \tab 229 \tab 32 \tab 3 \tab 356 \tab 205 \tab 444 \tab - \tab 1372 \cr
   1991 \tab 98 \tab 274 \tab 28 \tab 3 \tab 435 \tab 186 \tab 493 \tab - \tab 1517 \cr
   1993 \tab 106 \tab 283 \tab 27 \tab 3 \tab 443 \tab 218 \tab 526 \tab - \tab 1606 \cr
   1994 \tab 192 \tab 496 \tab 60 \tab 10 \tab 830 \tab 389 \tab 1015 \tab - \tab 2992 \cr
   1996 \tab 183 \tab 452 \tab 58 \tab 6 \tab 832 \tab 429 \tab 944 \tab - \tab 2904 \cr
   1998 \tab 166 \tab 476 \tab 65 \tab 2 \tab 784 \tab 378 \tab 961 \tab - \tab 2832 \cr
   2000 \tab 157 \tab 502 \tab 75 \tab 10 \tab 749 \tab 384 \tab 940 \tab - \tab 2817 \cr
   2002 \tab 73 \tab 242 \tab 18 \tab 5 \tab 394 \tab 176 \tab 1857 \tab - \tab 2765 \cr
   2004 \tab 79 \tab 238 \tab 18 \tab 5 \tab 409 \tab 157 \tab 1906 \tab - \tab 2812 \cr
   2006 \tab 167 \tab 509 \tab 44 \tab 11 \tab 882 \tab 379 \tab 2518 \tab - \tab 4510 \cr
   2008 \tab 90 \tab 305 \tab 30 \tab 6 \tab 692 \tab 206 \tab 694 \tab - \tab 2023 \cr
   2010 \tab 116 \tab 305 \tab 21 \tab 5 \tab 742 \tab 241 \tab 614 \tab - \tab 2044 \cr
   2012 \tab 72 \tab 273 \tab 33 \tab 14 \tab 704 \tab 206 \tab 672 \tab - \tab 1974 \cr
   2014 \tab 111 \tab 324 \tab 18 \tab 9 \tab 956 \tab 257 \tab 863 \tab - \tab 2538 \cr
   2016 \tab 100 \tab 389 \tab 25 \tab 10 \tab 1101 \tab 263 \tab 979 \tab - \tab 2867 \cr
   2018 \tab 95 \tab 278 \tab 19 \tab 11 \tab 943 \tab 213 \tab 789 \tab - \tab 2348 \cr
   2021 \tab 162 \tab 390 \tab 9 \tab 2 \tab 1761 \tab 367 \tab 1315 \tab 26 \tab 4032 \cr
   2022 \tab 104 \tab 308 \tab 36 \tab 8 \tab 1607 \tab 305 \tab 1168 \tab 8 \tab 3544 \cr
   Total \tab 3772 \tab 10933 \tab 1091 \tab 198 \tab 20428 \tab 8449 \tab 27485 \tab 34 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} always wrong
\item \code{2} almost always wrong
\item \code{3} wrong only sometimes
\item \code{4} not wrong at all
\item \code{5} other
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
