% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_19.R
\name{pres68}
\alias{pres68}
\title{pres68}
\source{
General Social Survey https://gss.norc.org
}
\description{
Vote for humphrey, nixon, or wallace
}
\details{
Question 57a. If voted: Did you vote for Humphrey, Nixon, or Wallace?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab don't know \tab humphrey \tab iap \tab nixon \tab no answer \tab other \tab refused \tab wallace \tab Total \cr
   1972 \tab 26 \tab 429 \tab 551 \tab 478 \tab 6 \tab 9 \tab 21 \tab 93 \tab 1613 \cr
   1973 \tab 17 \tab 358 \tab 574 \tab 407 \tab 3 \tab 7 \tab 11 \tab 127 \tab 1504 \cr
   1974 \tab - \tab - \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab - \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab - \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab - \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab - \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab - \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab - \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab - \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab - \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab - \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab - \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab - \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab - \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 43 \tab 787 \tab 70398 \tab 885 \tab 9 \tab 16 \tab 32 \tab 220 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} humphrey
\item \code{2} nixon
\item \code{3} wallace
\item \code{4} other
\item \code{5} refused
\item \code{6} wldnt vt-relig
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
