% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_03.R
\name{prestg10}
\alias{prestg10}
\title{prestg10}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's occupational prestige score (2010)
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 16 \tab 17 \tab 18 \tab 20 \tab 21 \tab 22 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 80 \tab iap \tab 23 \tab 76 \tab Total \cr
   1972 \tab 3 \tab 6 \tab 1 \tab 1 \tab 6 \tab 12 \tab 38 \tab 87 \tab 5 \tab 21 \tab 39 \tab 15 \tab 2 \tab 71 \tab 61 \tab 34 \tab 13 \tab 94 \tab 31 \tab 24 \tab 56 \tab 33 \tab 19 \tab 5 \tab 19 \tab 16 \tab 22 \tab 63 \tab 59 \tab 96 \tab 29 \tab 25 \tab 45 \tab 9 \tab 3 \tab 19 \tab 3 \tab 2 \tab 4 \tab 1 \tab 5 \tab 15 \tab 19 \tab 34 \tab 16 \tab 5 \tab 44 \tab 5 \tab 5 \tab 2 \tab 3 \tab 11 \tab 2 \tab 2 \tab 4 \tab 2 \tab 11 \tab 1 \tab 1 \tab 334 \tab - \tab - \tab 1613 \cr
   1973 \tab 1 \tab 5 \tab 2 \tab 3 \tab 12 \tab 11 \tab 37 \tab 55 \tab 10 \tab 29 \tab 43 \tab 10 \tab 8 \tab 78 \tab 49 \tab 24 \tab 14 \tab 96 \tab 30 \tab 18 \tab 45 \tab 38 \tab 35 \tab 6 \tab 17 \tab 20 \tab 32 \tab 80 \tab 39 \tab 101 \tab 39 \tab 25 \tab 30 \tab 14 \tab 2 \tab 18 \tab 4 \tab 3 \tab 3 \tab 3 \tab 3 \tab 16 \tab 13 \tab 31 \tab 9 \tab 5 \tab 37 \tab 11 \tab 2 \tab 1 \tab - \tab 7 \tab 7 \tab - \tab 8 \tab 3 \tab 14 \tab 1 \tab 1 \tab 245 \tab 1 \tab - \tab 1504 \cr
   1974 \tab 3 \tab 6 \tab 1 \tab 5 \tab 9 \tab 7 \tab 29 \tab 70 \tab 11 \tab 22 \tab 38 \tab 12 \tab 5 \tab 71 \tab 48 \tab 29 \tab 11 \tab 75 \tab 27 \tab 34 \tab 50 \tab 41 \tab 25 \tab 6 \tab 20 \tab 21 \tab 32 \tab 67 \tab 48 \tab 97 \tab 43 \tab 20 \tab 31 \tab 8 \tab 4 \tab 19 \tab 7 \tab 4 \tab 2 \tab 1 \tab 2 \tab 17 \tab 12 \tab 44 \tab 10 \tab 11 \tab 48 \tab 1 \tab 6 \tab 2 \tab - \tab 13 \tab 4 \tab 2 \tab 4 \tab 8 \tab 11 \tab 1 \tab 3 \tab 225 \tab 1 \tab - \tab 1484 \cr
   1975 \tab 2 \tab 9 \tab 1 \tab 2 \tab 9 \tab 9 \tab 34 \tab 72 \tab 1 \tab 26 \tab 43 \tab 9 \tab 2 \tab 65 \tab 62 \tab 24 \tab 8 \tab 77 \tab 32 \tab 18 \tab 61 \tab 38 \tab 32 \tab 5 \tab 20 \tab 12 \tab 33 \tab 83 \tab 40 \tab 91 \tab 43 \tab 31 \tab 20 \tab 13 \tab 3 \tab 20 \tab 2 \tab 6 \tab 1 \tab 2 \tab 4 \tab 7 \tab 21 \tab 33 \tab 7 \tab 11 \tab 40 \tab 3 \tab 11 \tab 2 \tab - \tab 16 \tab 5 \tab 1 \tab 5 \tab 10 \tab 5 \tab 1 \tab - \tab 246 \tab 1 \tab - \tab 1490 \cr
   1976 \tab 1 \tab 5 \tab 2 \tab 3 \tab 13 \tab 12 \tab 26 \tab 72 \tab 12 \tab 27 \tab 32 \tab 6 \tab 5 \tab 83 \tab 63 \tab 29 \tab 11 \tab 98 \tab 29 \tab 28 \tab 69 \tab 34 \tab 31 \tab 4 \tab 23 \tab 9 \tab 18 \tab 69 \tab 54 \tab 97 \tab 37 \tab 17 \tab 44 \tab 9 \tab 4 \tab 21 \tab 9 \tab 7 \tab 5 \tab 4 \tab 1 \tab 13 \tab 17 \tab 31 \tab 6 \tab 9 \tab 38 \tab 6 \tab 7 \tab - \tab - \tab 14 \tab 5 \tab 3 \tab 11 \tab 7 \tab 10 \tab 1 \tab 2 \tab 196 \tab - \tab - \tab 1499 \cr
   1977 \tab 5 \tab 5 \tab 3 \tab 1 \tab 3 \tab 18 \tab 40 \tab 65 \tab 5 \tab 27 \tab 45 \tab 13 \tab 7 \tab 70 \tab 59 \tab 25 \tab 13 \tab 94 \tab 30 \tab 14 \tab 54 \tab 39 \tab 37 \tab 4 \tab 20 \tab 12 \tab 28 \tab 56 \tab 49 \tab 90 \tab 41 \tab 24 \tab 27 \tab 7 \tab 2 \tab 15 \tab 2 \tab 6 \tab 1 \tab 1 \tab 4 \tab 10 \tab 15 \tab 40 \tab 11 \tab 12 \tab 47 \tab 11 \tab 11 \tab - \tab - \tab 9 \tab 3 \tab 3 \tab 8 \tab 4 \tab 11 \tab 1 \tab 2 \tab 270 \tab 1 \tab - \tab 1530 \cr
   1978 \tab 3 \tab 4 \tab 1 \tab 3 \tab 17 \tab 8 \tab 35 \tab 66 \tab 9 \tab 23 \tab 28 \tab 12 \tab 12 \tab 94 \tab 60 \tab 26 \tab 11 \tab 90 \tab 39 \tab 33 \tab 79 \tab 42 \tab 36 \tab 4 \tab 19 \tab 9 \tab 38 \tab 73 \tab 41 \tab 113 \tab 45 \tab 33 \tab 35 \tab 9 \tab 7 \tab 9 \tab 4 \tab 5 \tab 5 \tab 6 \tab 6 \tab 13 \tab 20 \tab 43 \tab 9 \tab 13 \tab 42 \tab 7 \tab 9 \tab 2 \tab - \tab 9 \tab 5 \tab 3 \tab 12 \tab 8 \tab 7 \tab - \tab 5 \tab 131 \tab 2 \tab - \tab 1532 \cr
   1980 \tab 4 \tab 8 \tab 1 \tab 2 \tab 10 \tab 2 \tab 31 \tab 49 \tab 10 \tab 22 \tab 31 \tab 10 \tab 5 \tab 68 \tab 66 \tab 33 \tab 9 \tab 80 \tab 34 \tab 22 \tab 74 \tab 53 \tab 40 \tab 9 \tab 21 \tab 11 \tab 22 \tab 66 \tab 56 \tab 99 \tab 45 \tab 26 \tab 33 \tab 13 \tab 2 \tab 23 \tab 5 \tab 2 \tab 3 \tab 4 \tab 8 \tab 16 \tab 18 \tab 37 \tab 7 \tab 18 \tab 38 \tab 5 \tab 4 \tab - \tab 1 \tab 14 \tab 2 \tab 1 \tab 9 \tab 7 \tab 8 \tab 1 \tab 3 \tab 167 \tab - \tab - \tab 1468 \cr
   1982 \tab 2 \tab 5 \tab 1 \tab 6 \tab 10 \tab 10 \tab 40 \tab 114 \tab 11 \tab 37 \tab 66 \tab 12 \tab 6 \tab 91 \tab 72 \tab 33 \tab 16 \tab 126 \tab 38 \tab 34 \tab 88 \tab 60 \tab 54 \tab 6 \tab 34 \tab 23 \tab 31 \tab 75 \tab 53 \tab 102 \tab 71 \tab 34 \tab 38 \tab 13 \tab 4 \tab 30 \tab 9 \tab 6 \tab 4 \tab 4 \tab 5 \tab 15 \tab 17 \tab 57 \tab 9 \tab 10 \tab 48 \tab 7 \tab 4 \tab - \tab - \tab 18 \tab 2 \tab - \tab 9 \tab 7 \tab 9 \tab 3 \tab 2 \tab 168 \tab 1 \tab - \tab 1860 \cr
   1983 \tab 3 \tab 1 \tab 4 \tab 2 \tab 12 \tab 11 \tab 28 \tab 65 \tab 7 \tab 23 \tab 48 \tab 7 \tab 5 \tab 74 \tab 52 \tab 41 \tab 14 \tab 86 \tab 31 \tab 33 \tab 80 \tab 45 \tab 45 \tab 9 \tab 17 \tab 18 \tab 28 \tab 81 \tab 60 \tab 118 \tab 57 \tab 30 \tab 32 \tab 13 \tab 7 \tab 19 \tab 12 \tab 13 \tab 7 \tab 2 \tab 6 \tab 11 \tab 26 \tab 43 \tab 10 \tab 11 \tab 44 \tab 9 \tab 6 \tab 2 \tab 2 \tab 9 \tab 5 \tab 1 \tab 10 \tab 8 \tab 19 \tab 1 \tab 2 \tab 133 \tab 1 \tab - \tab 1599 \cr
   1984 \tab 3 \tab 4 \tab - \tab 3 \tab 15 \tab 12 \tab 33 \tab 62 \tab 11 \tab 19 \tab 42 \tab 13 \tab 7 \tab 97 \tab 42 \tab 40 \tab 12 \tab 92 \tab 25 \tab 22 \tab 78 \tab 47 \tab 27 \tab 8 \tab 25 \tab 17 \tab 25 \tab 62 \tab 56 \tab 86 \tab 54 \tab 24 \tab 32 \tab 13 \tab 12 \tab 30 \tab 8 \tab 8 \tab 4 \tab 3 \tab 2 \tab 7 \tab 21 \tab 23 \tab 9 \tab 17 \tab 54 \tab 9 \tab 5 \tab 1 \tab - \tab 15 \tab 3 \tab 1 \tab 13 \tab 8 \tab 8 \tab 1 \tab 5 \tab 95 \tab 3 \tab - \tab 1473 \cr
   1985 \tab 1 \tab 7 \tab 4 \tab 2 \tab 7 \tab 13 \tab 28 \tab 59 \tab 6 \tab 25 \tab 38 \tab 8 \tab 11 \tab 88 \tab 53 \tab 28 \tab 12 \tab 80 \tab 21 \tab 33 \tab 82 \tab 47 \tab 17 \tab 7 \tab 21 \tab 18 \tab 25 \tab 78 \tab 64 \tab 110 \tab 44 \tab 39 \tab 33 \tab 20 \tab 3 \tab 34 \tab 3 \tab 7 \tab 6 \tab 7 \tab 5 \tab 14 \tab 26 \tab 35 \tab 6 \tab 11 \tab 46 \tab 12 \tab 5 \tab 1 \tab - \tab 14 \tab 4 \tab 7 \tab 14 \tab 7 \tab 9 \tab 2 \tab 6 \tab 120 \tab 1 \tab - \tab 1534 \cr
   1986 \tab 4 \tab 3 \tab 3 \tab - \tab 13 \tab 11 \tab 29 \tab 55 \tab 13 \tab 21 \tab 31 \tab 8 \tab 5 \tab 84 \tab 46 \tab 29 \tab 13 \tab 88 \tab 40 \tab 25 \tab 80 \tab 59 \tab 26 \tab 11 \tab 20 \tab 17 \tab 29 \tab 67 \tab 47 \tab 89 \tab 55 \tab 28 \tab 35 \tab 18 \tab 3 \tab 20 \tab 6 \tab 8 \tab 3 \tab 2 \tab 7 \tab 13 \tab 21 \tab 32 \tab 8 \tab 8 \tab 49 \tab 9 \tab 1 \tab - \tab - \tab 12 \tab 5 \tab 1 \tab 9 \tab 10 \tab 18 \tab 3 \tab 7 \tab 112 \tab 1 \tab - \tab 1470 \cr
   1987 \tab 4 \tab 7 \tab 3 \tab 2 \tab 11 \tab 13 \tab 42 \tab 111 \tab 8 \tab 26 \tab 51 \tab 16 \tab 4 \tab 87 \tab 48 \tab 38 \tab 19 \tab 89 \tab 51 \tab 25 \tab 86 \tab 57 \tab 34 \tab 12 \tab 22 \tab 18 \tab 36 \tab 67 \tab 64 \tab 101 \tab 82 \tab 35 \tab 51 \tab 11 \tab 7 \tab 26 \tab 16 \tab 9 \tab 4 \tab 3 \tab 5 \tab 11 \tab 22 \tab 33 \tab 10 \tab 13 \tab 57 \tab 11 \tab 6 \tab 1 \tab - \tab 20 \tab 9 \tab 2 \tab 11 \tab 9 \tab 13 \tab 4 \tab 4 \tab 177 \tab 5 \tab - \tab 1819 \cr
   1988 \tab 4 \tab 8 \tab 3 \tab 1 \tab 5 \tab 8 \tab 29 \tab 71 \tab 5 \tab 14 \tab 41 \tab 9 \tab 6 \tab 94 \tab 53 \tab 39 \tab 14 \tab 100 \tab 30 \tab 32 \tab 85 \tab 36 \tab 33 \tab 7 \tab 15 \tab 12 \tab 26 \tab 83 \tab 45 \tab 91 \tab 49 \tab 30 \tab 28 \tab 15 \tab 8 \tab 27 \tab 7 \tab 7 \tab 3 \tab 3 \tab 4 \tab 10 \tab 25 \tab 31 \tab 8 \tab 15 \tab 40 \tab 12 \tab 8 \tab 3 \tab - \tab 11 \tab 8 \tab 1 \tab 8 \tab 10 \tab 13 \tab 2 \tab 6 \tab 89 \tab 1 \tab - \tab 1481 \cr
   1989 \tab 5 \tab 5 \tab 1 \tab 1 \tab 10 \tab 9 \tab 30 \tab 57 \tab 8 \tab 17 \tab 40 \tab 17 \tab 6 \tab 91 \tab 40 \tab 31 \tab 9 \tab 93 \tab 26 \tab 26 \tab 95 \tab 48 \tab 22 \tab 9 \tab 21 \tab 22 \tab 29 \tab 56 \tab 58 \tab 101 \tab 53 \tab 29 \tab 35 \tab 15 \tab 7 \tab 29 \tab 7 \tab 9 \tab 8 \tab 6 \tab 12 \tab 15 \tab 21 \tab 30 \tab 12 \tab 12 \tab 53 \tab 7 \tab 4 \tab 2 \tab - \tab 22 \tab 7 \tab - \tab 12 \tab 13 \tab 17 \tab 2 \tab 4 \tab 109 \tab 2 \tab - \tab 1537 \cr
   1990 \tab 4 \tab 5 \tab 5 \tab 3 \tab 9 \tab 6 \tab 21 \tab 44 \tab 4 \tab 19 \tab 36 \tab 17 \tab 6 \tab 61 \tab 47 \tab 32 \tab 8 \tab 101 \tab 20 \tab 18 \tab 75 \tab 54 \tab 20 \tab 8 \tab 18 \tab 18 \tab 24 \tab 67 \tab 53 \tab 93 \tab 52 \tab 43 \tab 32 \tab 13 \tab 12 \tab 22 \tab 11 \tab 7 \tab 5 \tab 8 \tab 4 \tab 15 \tab 20 \tab 30 \tab 8 \tab 10 \tab 41 \tab 7 \tab 4 \tab 1 \tab - \tab 11 \tab 3 \tab 3 \tab 6 \tab 15 \tab 13 \tab - \tab 4 \tab 73 \tab 3 \tab - \tab 1372 \cr
   1991 \tab 4 \tab 4 \tab 6 \tab 1 \tab 7 \tab 7 \tab 33 \tab 41 \tab 9 \tab 23 \tab 36 \tab 15 \tab 8 \tab 86 \tab 68 \tab 37 \tab 14 \tab 85 \tab 29 \tab 36 \tab 86 \tab 56 \tab 27 \tab 3 \tab 18 \tab 22 \tab 34 \tab 66 \tab 42 \tab 86 \tab 63 \tab 33 \tab 28 \tab 25 \tab 9 \tab 37 \tab 8 \tab 11 \tab 2 \tab 3 \tab 7 \tab 15 \tab 22 \tab 40 \tab 2 \tab 14 \tab 41 \tab 10 \tab 3 \tab 1 \tab 1 \tab 15 \tab 7 \tab 1 \tab 7 \tab 6 \tab 12 \tab - \tab 4 \tab 100 \tab 1 \tab - \tab 1517 \cr
   1993 \tab 3 \tab 6 \tab 3 \tab 2 \tab 8 \tab 6 \tab 31 \tab 63 \tab 7 \tab 15 \tab 44 \tab 9 \tab 9 \tab 79 \tab 53 \tab 34 \tab 10 \tab 108 \tab 38 \tab 27 \tab 90 \tab 47 \tab 21 \tab 15 \tab 26 \tab 19 \tab 32 \tab 72 \tab 63 \tab 96 \tab 60 \tab 39 \tab 39 \tab 17 \tab 18 \tab 35 \tab 9 \tab 7 \tab 3 \tab 3 \tab 12 \tab 18 \tab 28 \tab 40 \tab 10 \tab 12 \tab 45 \tab 16 \tab 7 \tab 4 \tab 1 \tab 20 \tab 5 \tab 6 \tab 12 \tab 4 \tab 7 \tab 1 \tab 7 \tab 85 \tab - \tab - \tab 1606 \cr
   1994 \tab 6 \tab 19 \tab 4 \tab 5 \tab 22 \tab 19 \tab 54 \tab 120 \tab 12 \tab 33 \tab 72 \tab 14 \tab 11 \tab 181 \tab 65 \tab 44 \tab 22 \tab 210 \tab 68 \tab 42 \tab 154 \tab 104 \tab 48 \tab 13 \tab 44 \tab 46 \tab 49 \tab 140 \tab 83 \tab 185 \tab 117 \tab 78 \tab 67 \tab 51 \tab 21 \tab 63 \tab 23 \tab 15 \tab 12 \tab 13 \tab 11 \tab 28 \tab 64 \tab 73 \tab 14 \tab 36 \tab 85 \tab 23 \tab 15 \tab 6 \tab 2 \tab 32 \tab 13 \tab 9 \tab 29 \tab 16 \tab 25 \tab 3 \tab 9 \tab 147 \tab 3 \tab - \tab 2992 \cr
   1996 \tab 3 \tab 9 \tab 10 \tab 4 \tab 21 \tab 18 \tab 69 \tab 118 \tab 14 \tab 33 \tab 61 \tab 17 \tab 10 \tab 158 \tab 81 \tab 62 \tab 31 \tab 211 \tab 66 \tab 42 \tab 176 \tab 104 \tab 50 \tab 12 \tab 22 \tab 40 \tab 53 \tab 124 \tab 88 \tab 141 \tab 121 \tab 60 \tab 94 \tab 47 \tab 22 \tab 68 \tab 24 \tab 19 \tab 11 \tab 14 \tab 13 \tab 22 \tab 52 \tab 73 \tab 10 \tab 21 \tab 89 \tab 26 \tab 19 \tab 2 \tab - \tab 27 \tab 10 \tab 3 \tab 29 \tab 19 \tab 15 \tab 5 \tab 16 \tab 123 \tab 2 \tab - \tab 2904 \cr
   1998 \tab 8 \tab 9 \tab 11 \tab - \tab 21 \tab 16 \tab 52 \tab 94 \tab 14 \tab 30 \tab 78 \tab 22 \tab 10 \tab 156 \tab 56 \tab 66 \tab 18 \tab 196 \tab 52 \tab 46 \tab 158 \tab 105 \tab 36 \tab 21 \tab 31 \tab 43 \tab 61 \tab 115 \tab 86 \tab 142 \tab 120 \tab 78 \tab 64 \tab 42 \tab 19 \tab 73 \tab 18 \tab 21 \tab 10 \tab 8 \tab 12 \tab 34 \tab 51 \tab 58 \tab 10 \tab 33 \tab 89 \tab 24 \tab 16 \tab 4 \tab 1 \tab 29 \tab 10 \tab 9 \tab 19 \tab 18 \tab 29 \tab 4 \tab 8 \tab 166 \tab 2 \tab - \tab 2832 \cr
   2000 \tab 5 \tab 14 \tab 9 \tab 3 \tab 20 \tab 20 \tab 67 \tab 111 \tab 21 \tab 27 \tab 85 \tab 19 \tab 10 \tab 156 \tab 68 \tab 60 \tab 26 \tab 177 \tab 44 \tab 34 \tab 160 \tab 95 \tab 48 \tab 20 \tab 35 \tab 42 \tab 50 \tab 110 \tab 85 \tab 123 \tab 123 \tab 67 \tab 67 \tab 45 \tab 25 \tab 73 \tab 15 \tab 31 \tab 6 \tab 10 \tab 13 \tab 24 \tab 70 \tab 77 \tab 11 \tab 25 \tab 88 \tab 31 \tab 9 \tab 2 \tab 2 \tab 29 \tab 8 \tab 4 \tab 19 \tab 15 \tab 29 \tab 2 \tab 5 \tab 143 \tab 4 \tab 1 \tab 2817 \cr
   2002 \tab 6 \tab 11 \tab 7 \tab 2 \tab 25 \tab 12 \tab 60 \tab 86 \tab 12 \tab 25 \tab 74 \tab 17 \tab 15 \tab 178 \tab 78 \tab 57 \tab 24 \tab 197 \tab 55 \tab 41 \tab 134 \tab 96 \tab 37 \tab 18 \tab 42 \tab 53 \tab 46 \tab 113 \tab 73 \tab 132 \tab 143 \tab 75 \tab 63 \tab 49 \tab 22 \tab 75 \tab 21 \tab 26 \tab 11 \tab 8 \tab 20 \tab 37 \tab 53 \tab 74 \tab 10 \tab 25 \tab 76 \tab 25 \tab 10 \tab 4 \tab - \tab 16 \tab 13 \tab 3 \tab 18 \tab 11 \tab 25 \tab 1 \tab 9 \tab 112 \tab 4 \tab - \tab 2765 \cr
   2004 \tab 6 \tab 12 \tab 7 \tab 7 \tab 19 \tab 16 \tab 54 \tab 76 \tab 10 \tab 22 \tab 71 \tab 19 \tab 14 \tab 162 \tab 65 \tab 48 \tab 25 \tab 168 \tab 41 \tab 27 \tab 137 \tab 99 \tab 56 \tab 14 \tab 40 \tab 37 \tab 57 \tab 124 \tab 74 \tab 136 \tab 151 \tab 59 \tab 57 \tab 50 \tab 25 \tab 81 \tab 20 \tab 29 \tab 14 \tab 16 \tab 19 \tab 41 \tab 68 \tab 87 \tab 21 \tab 30 \tab 72 \tab 25 \tab 18 \tab 3 \tab 1 \tab 35 \tab 10 \tab 3 \tab 26 \tab 15 \tab 22 \tab 3 \tab 13 \tab 154 \tab 1 \tab - \tab 2812 \cr
   2006 \tab 9 \tab 17 \tab 9 \tab 13 \tab 23 \tab 25 \tab 92 \tab 166 \tab 12 \tab 39 \tab 143 \tab 24 \tab 18 \tab 255 \tab 102 \tab 117 \tab 43 \tab 298 \tab 97 \tab 69 \tab 225 \tab 154 \tab 79 \tab 15 \tab 65 \tab 63 \tab 68 \tab 194 \tab 141 \tab 202 \tab 196 \tab 120 \tab 109 \tab 54 \tab 35 \tab 109 \tab 39 \tab 37 \tab 16 \tab 6 \tab 23 \tab 44 \tab 106 \tab 151 \tab 24 \tab 54 \tab 108 \tab 37 \tab 26 \tab 9 \tab - \tab 37 \tab 24 \tab 9 \tab 25 \tab 29 \tab 47 \tab 14 \tab 18 \tab 223 \tab 4 \tab - \tab 4510 \cr
   2008 \tab 6 \tab 10 \tab 6 \tab 4 \tab 17 \tab 5 \tab 32 \tab 60 \tab 17 \tab 26 \tab 57 \tab 12 \tab 8 \tab 114 \tab 54 \tab 57 \tab 22 \tab 137 \tab 40 \tab 29 \tab 96 \tab 68 \tab 31 \tab 11 \tab 24 \tab 23 \tab 50 \tab 78 \tab 52 \tab 86 \tab 93 \tab 42 \tab 39 \tab 44 \tab 11 \tab 81 \tab 20 \tab 18 \tab 5 \tab 7 \tab 14 \tab 23 \tab 52 \tab 56 \tab 11 \tab 20 \tab 56 \tab 19 \tab 10 \tab 3 \tab 3 \tab 18 \tab 7 \tab 5 \tab 17 \tab 11 \tab 23 \tab 7 \tab 3 \tab 72 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 7 \tab 16 \tab 9 \tab - \tab 19 \tab 10 \tab 51 \tab 69 \tab 7 \tab 15 \tab 57 \tab 16 \tab 11 \tab 128 \tab 41 \tab 49 \tab 16 \tab 129 \tab 35 \tab 30 \tab 96 \tab 67 \tab 26 \tab 14 \tab 22 \tab 25 \tab 36 \tab 67 \tab 62 \tab 79 \tab 101 \tab 36 \tab 63 \tab 33 \tab 11 \tab 58 \tab 11 \tab 26 \tab 6 \tab 4 \tab 9 \tab 24 \tab 46 \tab 72 \tab 14 \tab 31 \tab 60 \tab 11 \tab 15 \tab 1 \tab 1 \tab 15 \tab 9 \tab 1 \tab 19 \tab 12 \tab 24 \tab 1 \tab 2 \tab 117 \tab 2 \tab - \tab 2044 \cr
   2012 \tab 3 \tab 10 \tab 4 \tab 3 \tab 24 \tab 11 \tab 53 \tab 64 \tab 11 \tab 15 \tab 63 \tab 8 \tab 8 \tab 110 \tab 35 \tab 39 \tab 20 \tab 138 \tab 34 \tab 33 \tab 87 \tab 69 \tab 30 \tab 11 \tab 20 \tab 20 \tab 31 \tab 56 \tab 48 \tab 79 \tab 103 \tab 37 \tab 40 \tab 38 \tab 18 \tab 55 \tab 12 \tab 18 \tab 8 \tab 6 \tab 7 \tab 30 \tab 53 \tab 59 \tab 10 \tab 29 \tab 85 \tab 19 \tab 7 \tab 2 \tab - \tab 10 \tab 10 \tab 1 \tab 10 \tab 31 \tab 24 \tab 3 \tab 7 \tab 103 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 11 \tab 9 \tab 10 \tab 2 \tab 36 \tab 10 \tab 59 \tab 96 \tab 11 \tab 23 \tab 74 \tab 18 \tab 10 \tab 139 \tab 71 \tab 52 \tab 12 \tab 170 \tab 57 \tab 26 \tab 122 \tab 69 \tab 42 \tab 12 \tab 29 \tab 34 \tab 46 \tab 80 \tab 66 \tab 97 \tab 138 \tab 54 \tab 51 \tab 43 \tab 28 \tab 89 \tab 18 \tab 31 \tab 12 \tab 7 \tab 11 \tab 28 \tab 67 \tab 72 \tab 25 \tab 26 \tab 67 \tab 25 \tab 16 \tab 7 \tab 2 \tab 25 \tab 12 \tab 7 \tab 22 \tab 11 \tab 22 \tab 2 \tab 13 \tab 111 \tab 3 \tab - \tab 2538 \cr
   2016 \tab 6 \tab 8 \tab 10 \tab 7 \tab 29 \tab 10 \tab 68 \tab 78 \tab 14 \tab 25 \tab 89 \tab 14 \tab 18 \tab 183 \tab 59 \tab 65 \tab 23 \tab 186 \tab 51 \tab 32 \tab 138 \tab 89 \tab 48 \tab 17 \tab 32 \tab 22 \tab 50 \tab 102 \tab 72 \tab 123 \tab 174 \tab 51 \tab 64 \tab 45 \tab 40 \tab 81 \tab 28 \tab 31 \tab 11 \tab 8 \tab 18 \tab 39 \tab 65 \tab 81 \tab 20 \tab 21 \tab 101 \tab 39 \tab 17 \tab 6 \tab - \tab 25 \tab 10 \tab 8 \tab 29 \tab 17 \tab 35 \tab 2 \tab 11 \tab 119 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 2 \tab 7 \tab 3 \tab 3 \tab 23 \tab 21 \tab 57 \tab 95 \tab 16 \tab 24 \tab 54 \tab 15 \tab 9 \tab 102 \tab 48 \tab 56 \tab 17 \tab 139 \tab 31 \tab 37 \tab 110 \tab 67 \tab 33 \tab 13 \tab 21 \tab 35 \tab 36 \tab 78 \tab 62 \tab 92 \tab 119 \tab 70 \tab 53 \tab 40 \tab 21 \tab 87 \tab 19 \tab 35 \tab 11 \tab 9 \tab 14 \tab 36 \tab 60 \tab 64 \tab 26 \tab 20 \tab 85 \tab 25 \tab 12 \tab 7 \tab 1 \tab 26 \tab 6 \tab 4 \tab 29 \tab 18 \tab 30 \tab 7 \tab 5 \tab 100 \tab 3 \tab - \tab 2348 \cr
   2021 \tab 2 \tab 24 \tab 6 \tab 6 \tab 26 \tab 13 \tab 79 \tab 91 \tab 13 \tab 21 \tab 93 \tab 19 \tab 23 \tab 207 \tab 106 \tab 75 \tab 32 \tab 192 \tab 55 \tab 33 \tab 188 \tab 159 \tab 48 \tab 20 \tab 51 \tab 39 \tab 50 \tab 132 \tab 96 \tab 154 \tab 189 \tab 100 \tab 71 \tab 101 \tab 44 \tab 129 \tab 47 \tab 39 \tab 24 \tab 19 \tab 29 \tab 84 \tab 154 \tab 164 \tab 55 \tab 53 \tab 163 \tab 51 \tab 40 \tab 10 \tab 3 \tab 47 \tab 32 \tab 9 \tab 30 \tab 23 \tab 70 \tab 12 \tab 24 \tab 159 \tab 4 \tab - \tab 4032 \cr
   2022 \tab 6 \tab 11 \tab 5 \tab 5 \tab 36 \tab 21 \tab 62 \tab 99 \tab 15 \tab 17 \tab 115 \tab 26 \tab 9 \tab 190 \tab 79 \tab 66 \tab 45 \tab 222 \tab 59 \tab 41 \tab 151 \tab 123 \tab 36 \tab 23 \tab 46 \tab 52 \tab 75 \tab 118 \tab 67 \tab 120 \tab 168 \tab 73 \tab 85 \tab 66 \tab 34 \tab 116 \tab 27 \tab 28 \tab 14 \tab 11 \tab 17 \tab 47 \tab 114 \tab 134 \tab 38 \tab 35 \tab 126 \tab 54 \tab 17 \tab 9 \tab 2 \tab 38 \tab 19 \tab 7 \tab 34 \tab 15 \tab 28 \tab 8 \tab 21 \tab 218 \tab 1 \tab - \tab 3544 \cr
   Total \tab 145 \tab 289 \tab 155 \tab 107 \tab 547 \tab 412 \tab 1523 \tab 2701 \tab 351 \tab 811 \tab 1958 \tab 478 \tab 303 \tab 3951 \tab 2050 \tab 1519 \tab 607 \tab 4522 \tab 1386 \tab 1064 \tab 3545 \tab 2342 \tab 1229 \tab 372 \tab 920 \tab 888 \tab 1302 \tab 2962 \tab 2146 \tab 3748 \tab 3018 \tab 1565 \tab 1635 \tab 1003 \tab 493 \tab 1691 \tab 474 \tab 531 \tab 244 \tab 212 \tab 332 \tab 792 \tab 1459 \tab 1982 \tab 466 \tab 676 \tab 2202 \tab 592 \tab 355 \tab 100 \tab 26 \tab 669 \tab 284 \tab 120 \tab 527 \tab 407 \tab 663 \tab 100 \tab 232 \tab 5142 \tab 64 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
