% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_03.R
\name{prestg105plus}
\alias{prestg105plus}
\title{R's occupational prestige score using threshold method (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
prestg105plus
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 6 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 68 \tab 69 \tab 7 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 76 \tab 77 \tab 78 \tab 79 \tab 8 \tab 80 \tab 81 \tab 83 \tab 84 \tab 85 \tab 86 \tab 87 \tab 88 \tab 89 \tab 9 \tab 90 \tab 91 \tab 92 \tab 93 \tab 94 \tab 95 \tab iap \tab 60 \tab 97 \tab 82 \tab 67 \tab not available in this release \tab Total \cr
   1972 \tab 8 \tab 4 \tab 15 \tab 15 \tab 57 \tab 25 \tab 67 \tab 8 \tab 44 \tab 15 \tab 11 \tab 22 \tab 21 \tab 14 \tab 23 \tab 11 \tab 54 \tab 21 \tab 30 \tab 85 \tab 2 \tab 4 \tab 14 \tab 6 \tab 6 \tab 19 \tab 1 \tab 12 \tab 18 \tab 13 \tab 9 \tab 1 \tab 4 \tab 5 \tab 6 \tab 8 \tab 12 \tab 4 \tab 27 \tab 21 \tab 1 \tab 36 \tab 11 \tab 4 \tab 13 \tab 22 \tab 82 \tab 6 \tab 3 \tab 40 \tab 6 \tab 8 \tab 8 \tab 30 \tab 14 \tab 7 \tab 11 \tab 4 \tab 9 \tab 13 \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab 17 \tab 1 \tab 8 \tab 1 \tab 4 \tab 12 \tab 2 \tab 3 \tab 6 \tab 41 \tab 5 \tab 26 \tab 11 \tab 1 \tab 3 \tab 3 \tab 28 \tab 1 \tab 3 \tab 19 \tab 1 \tab 334 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 3 \tab 10 \tab 30 \tab 13 \tab 33 \tab 29 \tab 58 \tab 7 \tab 62 \tab 7 \tab 12 \tab 19 \tab 22 \tab 13 \tab 15 \tab 8 \tab 41 \tab 33 \tab 36 \tab 68 \tab - \tab 4 \tab 14 \tab 4 \tab 4 \tab 15 \tab 1 \tab 19 \tab 22 \tab 16 \tab 7 \tab - \tab 10 \tab 9 \tab 7 \tab 4 \tab 15 \tab 8 \tab 39 \tab 25 \tab 1 \tab 42 \tab 7 \tab 2 \tab 11 \tab 21 \tab 104 \tab 9 \tab - \tab 23 \tab 10 \tab 6 \tab 12 \tab 18 \tab 15 \tab 4 \tab 8 \tab 7 \tab 6 \tab 6 \tab 3 \tab - \tab 1 \tab 1 \tab 8 \tab 2 \tab 5 \tab 7 \tab 1 \tab 7 \tab 3 \tab 5 \tab 13 \tab 2 \tab 4 \tab 5 \tab 32 \tab 2 \tab 15 \tab 3 \tab 7 \tab 7 \tab 10 \tab 33 \tab 1 \tab 1 \tab 21 \tab 4 \tab 245 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 3 \tab 7 \tab 37 \tab 11 \tab 35 \tab 25 \tab 66 \tab 1 \tab 44 \tab 9 \tab 8 \tab 22 \tab 24 \tab 12 \tab 21 \tab 10 \tab 26 \tab 31 \tab 31 \tab 66 \tab 1 \tab 1 \tab 9 \tab 8 \tab 9 \tab 12 \tab 5 \tab 10 \tab 28 \tab 24 \tab 5 \tab 1 \tab 10 \tab 5 \tab 9 \tab 6 \tab 13 \tab 2 \tab 34 \tab 25 \tab 3 \tab 57 \tab 9 \tab 1 \tab 11 \tab 21 \tab 79 \tab 10 \tab 2 \tab 27 \tab 10 \tab 7 \tab 10 \tab 24 \tab 7 \tab 8 \tab 12 \tab 8 \tab 12 \tab 3 \tab 7 \tab 1 \tab 1 \tab 2 \tab 7 \tab - \tab 3 \tab 10 \tab 1 \tab 8 \tab 2 \tab 5 \tab 11 \tab 1 \tab 1 \tab 7 \tab 52 \tab 6 \tab 18 \tab 9 \tab 5 \tab 2 \tab 3 \tab 37 \tab 2 \tab - \tab 20 \tab 5 \tab 225 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 1484 \cr
   1975 \tab 6 \tab 3 \tab 33 \tab 10 \tab 40 \tab 24 \tab 64 \tab 5 \tab 42 \tab 15 \tab 5 \tab 16 \tab 39 \tab 14 \tab 16 \tab 4 \tab 36 \tab 28 \tab 26 \tab 67 \tab 1 \tab 2 \tab 10 \tab 11 \tab 5 \tab 11 \tab 7 \tab 18 \tab 25 \tab 22 \tab 5 \tab 1 \tab 6 \tab 3 \tab 7 \tab 3 \tab 19 \tab 6 \tab 38 \tab 25 \tab 2 \tab 58 \tab 12 \tab 4 \tab 7 \tab 26 \tab 86 \tab 10 \tab 2 \tab 22 \tab 3 \tab 9 \tab 8 \tab 16 \tab 13 \tab 3 \tab 16 \tab 9 \tab 5 \tab 4 \tab 2 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 5 \tab 7 \tab - \tab 9 \tab 1 \tab 3 \tab 7 \tab 1 \tab 1 \tab 8 \tab 43 \tab 12 \tab 22 \tab 16 \tab 4 \tab 6 \tab 6 \tab 21 \tab 5 \tab 1 \tab 18 \tab 1 \tab 246 \tab 1 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 4 \tab 12 \tab 33 \tab 10 \tab 38 \tab 22 \tab 62 \tab 7 \tab 52 \tab 4 \tab 8 \tab 25 \tab 27 \tab 13 \tab 22 \tab 10 \tab 48 \tab 27 \tab 38 \tab 73 \tab 1 \tab 3 \tab 14 \tab 5 \tab 4 \tab 14 \tab 6 \tab 12 \tab 18 \tab 35 \tab 9 \tab - \tab 14 \tab 7 \tab 6 \tab 5 \tab 8 \tab 7 \tab 41 \tab 11 \tab 1 \tab 48 \tab 4 \tab 4 \tab 8 \tab 21 \tab 82 \tab 3 \tab - \tab 39 \tab 4 \tab 5 \tab 11 \tab 28 \tab 14 \tab 9 \tab 13 \tab 8 \tab 8 \tab 8 \tab 4 \tab 4 \tab 2 \tab 1 \tab 10 \tab 3 \tab 2 \tab 7 \tab - \tab 10 \tab 2 \tab 10 \tab 9 \tab 3 \tab 2 \tab 4 \tab 39 \tab 6 \tab 19 \tab 12 \tab 5 \tab 5 \tab 11 \tab 30 \tab 3 \tab - \tab 18 \tab 5 \tab 196 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1499 \cr
   1977 \tab 2 \tab 2 \tab 26 \tab 21 \tab 36 \tab 21 \tab 80 \tab 6 \tab 46 \tab 12 \tab 14 \tab 14 \tab 26 \tab 19 \tab 20 \tab 3 \tab 53 \tab 28 \tab 27 \tab 63 \tab 1 \tab 2 \tab 10 \tab 9 \tab 5 \tab 19 \tab 9 \tab 15 \tab 22 \tab 22 \tab 10 \tab 2 \tab 13 \tab 8 \tab 6 \tab 7 \tab 9 \tab - \tab 29 \tab 26 \tab 5 \tab 48 \tab 4 \tab 5 \tab 6 \tab 19 \tab 79 \tab 9 \tab 4 \tab 30 \tab 6 \tab 8 \tab 5 \tab 16 \tab 12 \tab 4 \tab 12 \tab 5 \tab 6 \tab 11 \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab 5 \tab 11 \tab 1 \tab 2 \tab 4 \tab 1 \tab 8 \tab - \tab - \tab 6 \tab 52 \tab 8 \tab 22 \tab 10 \tab 2 \tab 6 \tab 10 \tab 37 \tab 2 \tab - \tab 20 \tab 3 \tab 270 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1530 \cr
   1978 \tab 5 \tab 11 \tab 40 \tab 10 \tab 31 \tab 21 \tab 68 \tab 8 \tab 50 \tab 16 \tab 14 \tab 14 \tab 31 \tab 8 \tab 26 \tab 8 \tab 34 \tab 35 \tab 38 \tab 78 \tab 1 \tab 5 \tab 18 \tab 23 \tab 7 \tab 17 \tab 3 \tab 20 \tab 28 \tab 26 \tab 8 \tab 1 \tab 9 \tab 4 \tab 5 \tab 5 \tab 12 \tab 10 \tab 42 \tab 29 \tab 2 \tab 40 \tab 4 \tab 3 \tab 12 \tab 32 \tab 102 \tab 8 \tab 3 \tab 29 \tab 6 \tab 5 \tab 10 \tab 22 \tab 15 \tab 4 \tab 9 \tab 8 \tab 5 \tab 9 \tab 3 \tab - \tab 2 \tab 6 \tab 3 \tab 3 \tab 3 \tab 12 \tab 1 \tab 14 \tab 1 \tab 6 \tab 15 \tab 1 \tab 6 \tab 10 \tab 49 \tab 4 \tab 22 \tab 6 \tab 4 \tab 7 \tab 10 \tab 34 \tab 8 \tab - \tab 13 \tab 9 \tab 131 \tab - \tab 2 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1 \tab 7 \tab 22 \tab 4 \tab 31 \tab 23 \tab 53 \tab 10 \tab 48 \tab 14 \tab 7 \tab 26 \tab 29 \tab 12 \tab 28 \tab 10 \tab 34 \tab 18 \tab 29 \tab 68 \tab 3 \tab 8 \tab 13 \tab 7 \tab 7 \tab 18 \tab 8 \tab 20 \tab 17 \tab 31 \tab 15 \tab - \tab 27 \tab 10 \tab 9 \tab 5 \tab 8 \tab 4 \tab 35 \tab 25 \tab 3 \tab 44 \tab 2 \tab 1 \tab 6 \tab 27 \tab 84 \tab 12 \tab 2 \tab 32 \tab 7 \tab 9 \tab 7 \tab 23 \tab 20 \tab 9 \tab 10 \tab 7 \tab 8 \tab 15 \tab 3 \tab - \tab - \tab 6 \tab 6 \tab - \tab 6 \tab 5 \tab - \tab 7 \tab 2 \tab 4 \tab 12 \tab 3 \tab 3 \tab 9 \tab 48 \tab 3 \tab 30 \tab 5 \tab 9 \tab 4 \tab 10 \tab 23 \tab 3 \tab 1 \tab 16 \tab 5 \tab 167 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 1468 \cr
   1982 \tab 2 \tab 13 \tab 45 \tab 10 \tab 82 \tab 30 \tab 78 \tab 5 \tab 58 \tab 22 \tab 18 \tab 19 \tab 29 \tab 20 \tab 30 \tab 10 \tab 56 \tab 46 \tab 54 \tab 89 \tab 1 \tab 4 \tab 21 \tab 13 \tab 6 \tab 25 \tab 6 \tab 20 \tab 27 \tab 42 \tab 11 \tab 2 \tab 27 \tab 14 \tab 9 \tab 8 \tab 11 \tab 11 \tab 37 \tab 23 \tab 2 \tab 78 \tab 7 \tab 4 \tab 17 \tab 29 \tab 87 \tab 14 \tab - \tab 34 \tab 7 \tab 4 \tab 8 \tab 26 \tab 21 \tab 6 \tab 10 \tab 8 \tab 14 \tab 8 \tab 6 \tab - \tab - \tab 3 \tab 13 \tab - \tab 8 \tab 8 \tab - \tab 8 \tab 3 \tab 3 \tab 15 \tab 4 \tab 9 \tab 11 \tab 54 \tab 5 \tab 30 \tab 11 \tab 10 \tab 7 \tab 11 \tab 25 \tab 4 \tab - \tab 15 \tab 3 \tab 168 \tab 4 \tab 1 \tab 3 \tab - \tab - \tab 1860 \cr
   1983 \tab 4 \tab 6 \tab 27 \tab 15 \tab 31 \tab 18 \tab 78 \tab 11 \tab 39 \tab 10 \tab 9 \tab 21 \tab 25 \tab 16 \tab 27 \tab 4 \tab 43 \tab 26 \tab 35 \tab 82 \tab 2 \tab 5 \tab 17 \tab 14 \tab 9 \tab 17 \tab 9 \tab 19 \tab 23 \tab 32 \tab 11 \tab 1 \tab 16 \tab 6 \tab 9 \tab 5 \tab 11 \tab 11 \tab 48 \tab 20 \tab 3 \tab 51 \tab 3 \tab 2 \tab 6 \tab 37 \tab 124 \tab 15 \tab 1 \tab 47 \tab 5 \tab 4 \tab 5 \tab 15 \tab 15 \tab 7 \tab 14 \tab 2 \tab 13 \tab 10 \tab 2 \tab 2 \tab 1 \tab 6 \tab 12 \tab 1 \tab 4 \tab 8 \tab 2 \tab 9 \tab 4 \tab 2 \tab 9 \tab 4 \tab 6 \tab 11 \tab 54 \tab 6 \tab 26 \tab 6 \tab 7 \tab 9 \tab 12 \tab 29 \tab 2 \tab 3 \tab 27 \tab 6 \tab 133 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1599 \cr
   1984 \tab 4 \tab 13 \tab 33 \tab 14 \tab 30 \tab 28 \tab 57 \tab 3 \tab 77 \tab 12 \tab 15 \tab 11 \tab 15 \tab 16 \tab 26 \tab 9 \tab 38 \tab 29 \tab 39 \tab 72 \tab 3 \tab 1 \tab 11 \tab 4 \tab 9 \tab 19 \tab 4 \tab 18 \tab 17 \tab 37 \tab 11 \tab - \tab 17 \tab 6 \tab 8 \tab 7 \tab 10 \tab 7 \tab 35 \tab 29 \tab 3 \tab 45 \tab 6 \tab 7 \tab 10 \tab 30 \tab 78 \tab 7 \tab 4 \tab 32 \tab 9 \tab 4 \tab 9 \tab 17 \tab 17 \tab 10 \tab 19 \tab 4 \tab 13 \tab 6 \tab 5 \tab 3 \tab 2 \tab 4 \tab 10 \tab 1 \tab 7 \tab 8 \tab - \tab 10 \tab 3 \tab 6 \tab 5 \tab 2 \tab 3 \tab 8 \tab 35 \tab 6 \tab 32 \tab 11 \tab 7 \tab 4 \tab 12 \tab 35 \tab 5 \tab 1 \tab 13 \tab 10 \tab 95 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 1473 \cr
   1985 \tab 2 \tab 13 \tab 28 \tab 12 \tab 28 \tab 21 \tab 64 \tab 6 \tab 47 \tab 15 \tab 13 \tab 20 \tab 20 \tab 14 \tab 30 \tab 5 \tab 37 \tab 24 \tab 34 \tab 84 \tab - \tab 4 \tab 11 \tab 13 \tab 2 \tab 14 \tab 6 \tab 18 \tab 14 \tab 29 \tab 6 \tab - \tab 17 \tab 15 \tab 8 \tab 7 \tab 7 \tab 7 \tab 46 \tab 22 \tab 1 \tab 47 \tab 7 \tab 4 \tab 9 \tab 44 \tab 94 \tab 11 \tab 2 \tab 37 \tab 9 \tab 9 \tab 13 \tab 24 \tab 8 \tab 6 \tab 19 \tab 11 \tab 6 \tab 14 \tab 3 \tab - \tab - \tab 3 \tab 9 \tab 5 \tab 12 \tab 4 \tab 1 \tab 10 \tab 4 \tab 3 \tab 11 \tab 4 \tab 5 \tab 10 \tab 41 \tab 7 \tab 28 \tab 12 \tab 10 \tab 6 \tab 13 \tab 34 \tab 3 \tab 1 \tab 25 \tab 6 \tab 120 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 1534 \cr
   1986 \tab 3 \tab 9 \tab 34 \tab 11 \tab 25 \tab 25 \tab 60 \tab 6 \tab 62 \tab 6 \tab 10 \tab 21 \tab 16 \tab 7 \tab 20 \tab 6 \tab 43 \tab 40 \tab 40 \tab 70 \tab 1 \tab 6 \tab 14 \tab 11 \tab 7 \tab 10 \tab 6 \tab 17 \tab 19 \tab 35 \tab 10 \tab 1 \tab 21 \tab 11 \tab 11 \tab 2 \tab 4 \tab 6 \tab 31 \tab 31 \tab 3 \tab 48 \tab 8 \tab 2 \tab 13 \tab 19 \tab 83 \tab 12 \tab 4 \tab 33 \tab 5 \tab 5 \tab 6 \tab 22 \tab 23 \tab 9 \tab 11 \tab 10 \tab 6 \tab 10 \tab 1 \tab 3 \tab 1 \tab 5 \tab 8 \tab 1 \tab 3 \tab 6 \tab 1 \tab 12 \tab 5 \tab 9 \tab 13 \tab 3 \tab 2 \tab 6 \tab 40 \tab 3 \tab 34 \tab 10 \tab 6 \tab 2 \tab 9 \tab 20 \tab 4 \tab - \tab 32 \tab 10 \tab 112 \tab 6 \tab - \tab 2 \tab 1 \tab - \tab 1470 \cr
   1987 \tab 5 \tab 14 \tab 39 \tab 12 \tab 63 \tab 32 \tab 80 \tab 6 \tab 63 \tab 7 \tab 17 \tab 16 \tab 20 \tab 19 \tab 29 \tab 12 \tab 46 \tab 42 \tab 40 \tab 74 \tab 3 \tab 6 \tab 7 \tab 9 \tab 7 \tab 14 \tab 11 \tab 17 \tab 22 \tab 48 \tab 11 \tab - \tab 28 \tab 7 \tab 12 \tab 6 \tab 9 \tab 8 \tab 35 \tab 34 \tab 4 \tab 80 \tab 10 \tab 3 \tab 15 \tab 30 \tab 90 \tab 14 \tab 3 \tab 42 \tab 7 \tab 7 \tab 7 \tab 27 \tab 18 \tab 6 \tab 16 \tab 9 \tab 16 \tab 9 \tab 2 \tab 4 \tab 3 \tab 5 \tab 11 \tab 3 \tab 7 \tab 7 \tab 1 \tab 10 \tab 1 \tab 2 \tab 10 \tab 5 \tab 4 \tab 8 \tab 39 \tab 11 \tab 29 \tab 19 \tab 7 \tab 11 \tab 13 \tab 39 \tab 6 \tab 2 \tab 25 \tab 6 \tab 177 \tab 3 \tab 3 \tab 2 \tab 1 \tab - \tab 1819 \cr
   1988 \tab 5 \tab 4 \tab 26 \tab 9 \tab 37 \tab 24 \tab 53 \tab 5 \tab 65 \tab 13 \tab 16 \tab 14 \tab 30 \tab 17 \tab 29 \tab 7 \tab 50 \tab 36 \tab 38 \tab 71 \tab 3 \tab 3 \tab 15 \tab 12 \tab 4 \tab 9 \tab 5 \tab 12 \tab 14 \tab 50 \tab 5 \tab - \tab 19 \tab 14 \tab 9 \tab 4 \tab 7 \tab 2 \tab 29 \tab 24 \tab 4 \tab 53 \tab 8 \tab - \tab 14 \tab 32 \tab 88 \tab 10 \tab 2 \tab 21 \tab 8 \tab 10 \tab 7 \tab 16 \tab 12 \tab 14 \tab 19 \tab 5 \tab 12 \tab 8 \tab 2 \tab 2 \tab 3 \tab 8 \tab 9 \tab 3 \tab 4 \tab 3 \tab 2 \tab 9 \tab 2 \tab 2 \tab 12 \tab 3 \tab 4 \tab 9 \tab 41 \tab 9 \tab 33 \tab 7 \tab 7 \tab 5 \tab 10 \tab 23 \tab 1 \tab 1 \tab 26 \tab 11 \tab 89 \tab 3 \tab 4 \tab 2 \tab - \tab - \tab 1481 \cr
   1989 \tab 3 \tab 8 \tab 32 \tab 8 \tab 24 \tab 30 \tab 64 \tab 3 \tab 66 \tab 13 \tab 13 \tab 12 \tab 11 \tab 10 \tab 22 \tab 8 \tab 41 \tab 24 \tab 34 \tab 87 \tab - \tab 4 \tab 9 \tab 10 \tab 4 \tab 22 \tab 3 \tab 15 \tab 20 \tab 42 \tab 7 \tab - \tab 22 \tab 12 \tab 11 \tab 4 \tab 10 \tab 8 \tab 27 \tab 24 \tab 5 \tab 53 \tab 6 \tab 5 \tab 7 \tab 37 \tab 102 \tab 12 \tab 4 \tab 25 \tab 3 \tab 4 \tab 6 \tab 17 \tab 15 \tab 11 \tab 19 \tab 8 \tab 14 \tab 9 \tab 1 \tab - \tab 1 \tab 5 \tab 13 \tab 4 \tab 6 \tab 9 \tab 2 \tab 8 \tab 8 \tab 3 \tab 10 \tab 8 \tab 11 \tab 10 \tab 33 \tab 8 \tab 29 \tab 12 \tab 13 \tab 9 \tab 10 \tab 33 \tab 7 \tab 3 \tab 28 \tab 9 \tab 109 \tab 2 \tab - \tab 4 \tab - \tab - \tab 1537 \cr
   1990 \tab 6 \tab 9 \tab 23 \tab 8 \tab 22 \tab 13 \tab 45 \tab 2 \tab 50 \tab 14 \tab 7 \tab 13 \tab 19 \tab 17 \tab 24 \tab 2 \tab 37 \tab 30 \tab 44 \tab 64 \tab - \tab 5 \tab 11 \tab 12 \tab 2 \tab 10 \tab 8 \tab 18 \tab 18 \tab 38 \tab 14 \tab 1 \tab 17 \tab 8 \tab 9 \tab 2 \tab 11 \tab 3 \tab 30 \tab 17 \tab 4 \tab 57 \tab 6 \tab 1 \tab 5 \tab 27 \tab 94 \tab 14 \tab 4 \tab 27 \tab 7 \tab 9 \tab 8 \tab 12 \tab 26 \tab 15 \tab 19 \tab 5 \tab 8 \tab 16 \tab 3 \tab 4 \tab 1 \tab 8 \tab 12 \tab 2 \tab 6 \tab 6 \tab 2 \tab 6 \tab 5 \tab 2 \tab 15 \tab 5 \tab 5 \tab 8 \tab 38 \tab 4 \tab 22 \tab 8 \tab 5 \tab 5 \tab 7 \tab 29 \tab 4 \tab - \tab 29 \tab 5 \tab 73 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 3 \tab 6 \tab 28 \tab 7 \tab 16 \tab 24 \tab 67 \tab 4 \tab 61 \tab 9 \tab 18 \tab 19 \tab 29 \tab 12 \tab 29 \tab 8 \tab 36 \tab 29 \tab 46 \tab 83 \tab 2 \tab 3 \tab 10 \tab 14 \tab 5 \tab 15 \tab 5 \tab 12 \tab 18 \tab 40 \tab 6 \tab - \tab 27 \tab 9 \tab 5 \tab 7 \tab 14 \tab 9 \tab 34 \tab 33 \tab 4 \tab 63 \tab 6 \tab 4 \tab 5 \tab 27 \tab 83 \tab 13 \tab 3 \tab 23 \tab 9 \tab 6 \tab 6 \tab 19 \tab 14 \tab 8 \tab 16 \tab 9 \tab 15 \tab 17 \tab 3 \tab 1 \tab 2 \tab 8 \tab 13 \tab 5 \tab 7 \tab 3 \tab 3 \tab 8 \tab 5 \tab 4 \tab 12 \tab 5 \tab 9 \tab 8 \tab 47 \tab 2 \tab 22 \tab 12 \tab 7 \tab 7 \tab 10 \tab 25 \tab 5 \tab 2 \tab 16 \tab 6 \tab 100 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab 1517 \cr
   1993 \tab 6 \tab 7 \tab 29 \tab 10 \tab 25 \tab 22 \tab 66 \tab 4 \tab 45 \tab 6 \tab 20 \tab 18 \tab 26 \tab 17 \tab 28 \tab 9 \tab 38 \tab 34 \tab 43 \tab 76 \tab 3 \tab 6 \tab 14 \tab 16 \tab 10 \tab 14 \tab 5 \tab 11 \tab 21 \tab 39 \tab 6 \tab - \tab 21 \tab 12 \tab 13 \tab 5 \tab 13 \tab 9 \tab 31 \tab 24 \tab 2 \tab 61 \tab 6 \tab 5 \tab 7 \tab 40 \tab 93 \tab 18 \tab 2 \tab 30 \tab 9 \tab 10 \tab 11 \tab 15 \tab 19 \tab 13 \tab 18 \tab 10 \tab 12 \tab 16 \tab 2 \tab - \tab 8 \tab 9 \tab 16 \tab 2 \tab 12 \tab 6 \tab 2 \tab 8 \tab 1 \tab 4 \tab 17 \tab 7 \tab 7 \tab 14 \tab 46 \tab 10 \tab 28 \tab 9 \tab 18 \tab 4 \tab 19 \tab 31 \tab 3 \tab 1 \tab 15 \tab 8 \tab 85 \tab 7 \tab 3 \tab 5 \tab - \tab - \tab 1606 \cr
   1994 \tab 7 \tab 19 \tab 61 \tab 17 \tab 52 \tab 42 \tab 128 \tab 8 \tab 102 \tab 14 \tab 32 \tab 37 \tab 25 \tab 26 \tab 30 \tab 19 \tab 58 \tab 69 \tab 98 \tab 129 \tab 5 \tab 4 \tab 25 \tab 19 \tab 14 \tab 30 \tab 11 \tab 23 \tab 33 \tab 77 \tab 19 \tab 2 \tab 49 \tab 20 \tab 21 \tab 12 \tab 17 \tab 16 \tab 60 \tab 48 \tab 4 \tab 129 \tab 22 \tab 4 \tab 15 \tab 60 \tab 188 \tab 14 \tab 7 \tab 39 \tab 20 \tab 22 \tab 10 \tab 39 \tab 34 \tab 26 \tab 27 \tab 26 \tab 24 \tab 21 \tab 8 \tab 5 \tab 5 \tab 16 \tab 25 \tab 9 \tab 17 \tab 14 \tab 4 \tab 39 \tab 3 \tab 7 \tab 28 \tab 8 \tab 7 \tab 20 \tab 106 \tab 19 \tab 53 \tab 22 \tab 18 \tab 14 \tab 39 \tab 62 \tab 8 \tab 2 \tab 46 \tab 14 \tab 147 \tab 9 \tab 4 \tab 4 \tab 1 \tab - \tab 2992 \cr
   1996 \tab 11 \tab 19 \tab 68 \tab 15 \tab 47 \tab 53 \tab 106 \tab 7 \tab 87 \tab 22 \tab 37 \tab 24 \tab 27 \tab 23 \tab 49 \tab 15 \tab 83 \tab 74 \tab 92 \tab 160 \tab 4 \tab 8 \tab 29 \tab 21 \tab 9 \tab 29 \tab 7 \tab 23 \tab 32 \tab 69 \tab 30 \tab - \tab 42 \tab 14 \tab 21 \tab 4 \tab 17 \tab 10 \tab 53 \tab 42 \tab 3 \tab 130 \tab 18 \tab 5 \tab 19 \tab 50 \tab 136 \tab 24 \tab 7 \tab 36 \tab 21 \tab 13 \tab 10 \tab 54 \tab 44 \tab 24 \tab 33 \tab 26 \tab 34 \tab 18 \tab 4 \tab 3 \tab 4 \tab 17 \tab 28 \tab 5 \tab 18 \tab 9 \tab 2 \tab 24 \tab 12 \tab 10 \tab 23 \tab 10 \tab 13 \tab 16 \tab 83 \tab 25 \tab 50 \tab 20 \tab 12 \tab 21 \tab 32 \tab 57 \tab 8 \tab - \tab 46 \tab 21 \tab 123 \tab 7 \tab 4 \tab 7 \tab 2 \tab - \tab 2904 \cr
   1998 \tab 3 \tab 17 \tab 69 \tab 19 \tab 39 \tab 38 \tab 116 \tab 5 \tab 88 \tab 13 \tab 30 \tab 24 \tab 15 \tab 26 \tab 50 \tab 13 \tab 79 \tab 63 \tab 81 \tab 134 \tab 4 \tab 7 \tab 22 \tab 22 \tab 14 \tab 34 \tab 12 \tab 19 \tab 29 \tab 62 \tab 26 \tab - \tab 40 \tab 19 \tab 31 \tab 12 \tab 15 \tab 11 \tab 35 \tab 57 \tab 5 \tab 114 \tab 15 \tab 5 \tab 21 \tab 60 \tab 159 \tab 18 \tab 9 \tab 34 \tab 10 \tab 15 \tab 13 \tab 38 \tab 36 \tab 17 \tab 38 \tab 19 \tab 16 \tab 30 \tab - \tab 7 \tab 6 \tab 23 \tab 32 \tab 7 \tab 17 \tab 5 \tab 1 \tab 37 \tab 8 \tab 12 \tab 36 \tab 8 \tab 20 \tab 15 \tab 60 \tab 14 \tab 83 \tab 17 \tab 20 \tab 17 \tab 23 \tab 44 \tab 7 \tab 2 \tab 55 \tab 11 \tab 166 \tab 4 \tab 3 \tab 5 \tab 2 \tab - \tab 2832 \cr
   2000 \tab 5 \tab 15 \tab 48 \tab 19 \tab 52 \tab 53 \tab 138 \tab 10 \tab 91 \tab 14 \tab 34 \tab 25 \tab 28 \tab 33 \tab 37 \tab 19 \tab 64 \tab 55 \tab 77 \tab 125 \tab 3 \tab 1 \tab 21 \tab 17 \tab 14 \tab 25 \tab 12 \tab 24 \tab 35 \tab 69 \tab 20 \tab 1 \tab 47 \tab 15 \tab 33 \tab 15 \tab 22 \tab 14 \tab 32 \tab 37 \tab 3 \tab 127 \tab 15 \tab 2 \tab 22 \tab 59 \tab 126 \tab 21 \tab 9 \tab 39 \tab 17 \tab 19 \tab 12 \tab 28 \tab 26 \tab 30 \tab 44 \tab 15 \tab 25 \tab 28 \tab 4 \tab 3 \tab 11 \tab 15 \tab 27 \tab 10 \tab 14 \tab 8 \tab 3 \tab 27 \tab 9 \tab 9 \tab 25 \tab 5 \tab 19 \tab 19 \tab 95 \tab 11 \tab 61 \tab 20 \tab 12 \tab 21 \tab 34 \tab 50 \tab 8 \tab 5 \tab 49 \tab 9 \tab 143 \tab 10 \tab 1 \tab 8 \tab 6 \tab - \tab 2817 \cr
   2002 \tab 6 \tab 21 \tab 41 \tab 11 \tab 35 \tab 49 \tab 116 \tab 8 \tab 101 \tab 12 \tab 49 \tab 31 \tab 15 \tab 26 \tab 48 \tab 20 \tab 77 \tab 57 \tab 98 \tab 126 \tab 3 \tab 7 \tab 18 \tab 23 \tab 7 \tab 22 \tab 10 \tab 26 \tab 30 \tab 56 \tab 22 \tab 3 \tab 50 \tab 24 \tab 31 \tab 7 \tab 22 \tab 8 \tab 37 \tab 42 \tab 5 \tab 112 \tab 15 \tab 5 \tab 17 \tab 59 \tab 153 \tab 24 \tab 5 \tab 37 \tab 19 \tab 11 \tab 8 \tab 33 \tab 34 \tab 31 \tab 33 \tab 19 \tab 30 \tab 30 \tab 3 \tab 2 \tab 6 \tab 22 \tab 28 \tab 15 \tab 21 \tab 11 \tab 7 \tab 27 \tab 10 \tab 10 \tab 38 \tab 8 \tab 15 \tab 14 \tab 80 \tab 11 \tab 45 \tab 10 \tab 12 \tab 17 \tab 24 \tab 56 \tab 10 \tab 2 \tab 44 \tab 10 \tab 112 \tab 6 \tab 3 \tab 10 \tab 1 \tab - \tab 2765 \cr
   2004 \tab 3 \tab 18 \tab 26 \tab 18 \tab 33 \tab 44 \tab 114 \tab 7 \tab 83 \tab 18 \tab 43 \tab 23 \tab 21 \tab 30 \tab 41 \tab 14 \tab 59 \tab 46 \tab 81 \tab 106 \tab 4 \tab 2 \tab 26 \tab 17 \tab 7 \tab 28 \tab 6 \tab 22 \tab 37 \tab 71 \tab 16 \tab - \tab 57 \tab 22 \tab 23 \tab 10 \tab 12 \tab 13 \tab 42 \tab 53 \tab 6 \tab 139 \tab 11 \tab 3 \tab 17 \tab 48 \tab 149 \tab 22 \tab 6 \tab 37 \tab 16 \tab 15 \tab 13 \tab 29 \tab 38 \tab 34 \tab 36 \tab 19 \tab 33 \tab 22 \tab 7 \tab 1 \tab 9 \tab 20 \tab 22 \tab 13 \tab 31 \tab 8 \tab 4 \tab 35 \tab 10 \tab 4 \tab 42 \tab 11 \tab 19 \tab 27 \tab 100 \tab 18 \tab 38 \tab 27 \tab 20 \tab 15 \tab 30 \tab 57 \tab 10 \tab 2 \tab 44 \tab 19 \tab 154 \tab 8 \tab 2 \tab 15 \tab 1 \tab - \tab 2812 \cr
   2006 \tab 8 \tab 22 \tab 67 \tab 24 \tab 74 \tab 75 \tab 178 \tab 10 \tab 132 \tab 26 \tab 69 \tab 45 \tab 25 \tab 70 \tab 89 \tab 20 \tab 115 \tab 86 \tab 144 \tab 239 \tab 2 \tab 5 \tab 38 \tab 37 \tab 19 \tab 33 \tab 23 \tab 41 \tab 43 \tab 109 \tab 20 \tab 2 \tab 81 \tab 31 \tab 37 \tab 17 \tab 27 \tab 16 \tab 59 \tab 62 \tab 9 \tab 204 \tab 25 \tab 6 \tab 26 \tab 83 \tab 212 \tab 46 \tab 14 \tab 54 \tab 32 \tab 21 \tab 25 \tab 46 \tab 47 \tab 38 \tab 74 \tab 23 \tab 50 \tab 43 \tab 13 \tab 9 \tab 12 \tab 24 \tab 35 \tab 18 \tab 31 \tab 19 \tab 7 \tab 62 \tab 8 \tab 10 \tab 40 \tab 8 \tab 30 \tab 20 \tab 171 \tab 29 \tab 78 \tab 21 \tab 32 \tab 22 \tab 41 \tab 85 \tab 12 \tab 3 \tab 88 \tab 30 \tab 223 \tab 9 \tab 10 \tab 11 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 5 \tab 18 \tab 24 \tab 6 \tab 24 \tab 26 \tab 84 \tab 8 \tab 57 \tab 17 \tab 39 \tab 24 \tab 20 \tab 27 \tab 42 \tab 7 \tab 39 \tab 49 \tab 66 \tab 102 \tab 3 \tab 2 \tab 13 \tab 13 \tab 11 \tab 16 \tab 6 \tab 15 \tab 27 \tab 44 \tab 12 \tab 1 \tab 31 \tab 16 \tab 18 \tab 3 \tab 7 \tab 10 \tab 13 \tab 41 \tab 6 \tab 82 \tab 16 \tab 6 \tab 9 \tab 34 \tab 102 \tab 15 \tab 3 \tab 27 \tab 9 \tab 14 \tab 10 \tab 24 \tab 22 \tab 24 \tab 24 \tab 15 \tab 29 \tab 13 \tab 4 \tab 3 \tab 2 \tab 18 \tab 28 \tab 11 \tab 23 \tab 8 \tab 3 \tab 24 \tab 5 \tab 2 \tab 29 \tab 6 \tab 20 \tab 9 \tab 60 \tab 9 \tab 38 \tab 10 \tab 10 \tab 11 \tab 22 \tab 40 \tab 8 \tab 4 \tab 45 \tab 6 \tab 72 \tab 8 \tab 3 \tab 11 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 4 \tab 18 \tab 28 \tab 12 \tab 26 \tab 38 \tab 99 \tab 1 \tab 66 \tab 11 \tab 43 \tab 10 \tab 13 \tab 20 \tab 34 \tab 11 \tab 47 \tab 37 \tab 66 \tab 79 \tab 2 \tab 5 \tab 14 \tab 22 \tab 8 \tab 20 \tab 10 \tab 18 \tab 14 \tab 40 \tab 15 \tab - \tab 38 \tab 7 \tab 28 \tab 7 \tab 10 \tab 6 \tab 20 \tab 27 \tab 6 \tab 78 \tab 15 \tab 3 \tab 20 \tab 42 \tab 82 \tab 22 \tab 2 \tab 31 \tab 6 \tab 22 \tab 8 \tab 17 \tab 14 \tab 17 \tab 39 \tab 17 \tab 19 \tab 19 \tab - \tab 3 \tab 7 \tab 11 \tab 24 \tab 14 \tab 18 \tab 12 \tab 3 \tab 22 \tab 6 \tab 4 \tab 23 \tab 8 \tab 24 \tab 18 \tab 73 \tab 21 \tab 38 \tab 8 \tab 10 \tab 10 \tab 20 \tab 25 \tab 6 \tab 1 \tab 41 \tab 8 \tab 117 \tab 7 \tab 2 \tab 6 \tab 1 \tab - \tab 2044 \cr
   2012 \tab - \tab 16 \tab 27 \tab 11 \tab 31 \tab 43 \tab 80 \tab 3 \tab 49 \tab 11 \tab 38 \tab 14 \tab 9 \tab 29 \tab 36 \tab 8 \tab 34 \tab 47 \tab 68 \tab 89 \tab 1 \tab 4 \tab 16 \tab 28 \tab 4 \tab 8 \tab 9 \tab 10 \tab 16 \tab 41 \tab 10 \tab - \tab 43 \tab 3 \tab 29 \tab 5 \tab 7 \tab 7 \tab 20 \tab 29 \tab 3 \tab 88 \tab 9 \tab 3 \tab 17 \tab 37 \tab 76 \tab 14 \tab 3 \tab 16 \tab 17 \tab 13 \tab 15 \tab 17 \tab 11 \tab 23 \tab 31 \tab 9 \tab 23 \tab 17 \tab 4 \tab 1 \tab 4 \tab 12 \tab 16 \tab 9 \tab 11 \tab 4 \tab 2 \tab 25 \tab 10 \tab 2 \tab 30 \tab 6 \tab 18 \tab 13 \tab 76 \tab 13 \tab 52 \tab 11 \tab 10 \tab 19 \tab 23 \tab 41 \tab 5 \tab 1 \tab 39 \tab 30 \tab 103 \tab 3 \tab 2 \tab 3 \tab 1 \tab - \tab 1974 \cr
   2014 \tab 2 \tab 33 \tab 44 \tab 12 \tab 42 \tab 46 \tab 106 \tab 6 \tab 86 \tab 6 \tab 42 \tab 29 \tab 8 \tab 35 \tab 41 \tab 9 \tab 59 \tab 56 \tab 74 \tab 115 \tab 7 \tab 5 \tab 21 \tab 21 \tab 13 \tab 7 \tab 10 \tab 16 \tab 24 \tab 64 \tab 15 \tab - \tab 33 \tab 18 \tab 18 \tab 7 \tab 11 \tab 10 \tab 27 \tab 34 \tab 8 \tab 120 \tab 16 \tab 7 \tab 18 \tab 55 \tab 92 \tab 23 \tab 4 \tab 19 \tab 22 \tab 18 \tab 15 \tab 21 \tab 21 \tab 21 \tab 51 \tab 11 \tab 29 \tab 32 \tab 3 \tab 1 \tab 9 \tab 18 \tab 30 \tab 17 \tab 21 \tab 32 \tab 3 \tab 31 \tab 11 \tab 5 \tab 19 \tab 9 \tab 23 \tab 16 \tab 87 \tab 13 \tab 39 \tab 18 \tab 17 \tab 16 \tab 33 \tab 51 \tab 2 \tab 3 \tab 48 \tab 16 \tab 111 \tab 5 \tab 4 \tab 10 \tab 2 \tab - \tab 2538 \cr
   2016 \tab - \tab 21 \tab 32 \tab 13 \tab 36 \tab 54 \tab 109 \tab 12 \tab 85 \tab 13 \tab 68 \tab 17 \tab 10 \tab 39 \tab 54 \tab 12 \tab 76 \tab 54 \tab 101 \tab 123 \tab 4 \tab 3 \tab 23 \tab 26 \tab 9 \tab 23 \tab 3 \tab 13 \tab 24 \tab 72 \tab 12 \tab - \tab 58 \tab 7 \tab 52 \tab 13 \tab 12 \tab 7 \tab 39 \tab 42 \tab 5 \tab 147 \tab 17 \tab 3 \tab 17 \tab 56 \tab 107 \tab 30 \tab 6 \tab 29 \tab 25 \tab 18 \tab 17 \tab 32 \tab 22 \tab 27 \tab 45 \tab 11 \tab 40 \tab 13 \tab 7 \tab 6 \tab 10 \tab 20 \tab 24 \tab 13 \tab 29 \tab 18 \tab 7 \tab 39 \tab 12 \tab 5 \tab 42 \tab 10 \tab 18 \tab 19 \tab 85 \tab 24 \tab 62 \tab 17 \tab 16 \tab 21 \tab 41 \tab 59 \tab 11 \tab - \tab 55 \tab 15 \tab 119 \tab 7 \tab 7 \tab 11 \tab - \tab - \tab 2867 \cr
   2018 \tab 1 \tab 29 \tab 45 \tab 20 \tab 33 \tab 39 \tab 90 \tab 9 \tab 71 \tab 12 \tab 24 \tab 12 \tab 8 \tab 28 \tab 49 \tab 8 \tab 38 \tab 50 \tab 63 \tab 110 \tab 5 \tab 2 \tab 24 \tab 23 \tab 10 \tab 21 \tab 2 \tab 18 \tab 23 \tab 38 \tab 11 \tab - \tab 26 \tab 16 \tab 21 \tab 4 \tab 10 \tab 8 \tab 16 \tab 40 \tab 1 \tab 105 \tab 21 \tab 4 \tab 23 \tab 43 \tab 79 \tab 21 \tab 2 \tab 31 \tab 25 \tab 7 \tab 17 \tab 15 \tab 40 \tab 23 \tab 53 \tab 12 \tab 25 \tab 19 \tab 3 \tab 2 \tab 5 \tab 13 \tab 35 \tab 15 \tab 24 \tab 12 \tab 3 \tab 29 \tab 8 \tab 4 \tab 33 \tab 12 \tab 19 \tab 28 \tab 65 \tab 10 \tab 63 \tab 17 \tab 21 \tab 17 \tab 35 \tab 48 \tab 10 \tab 1 \tab 50 \tab 11 \tab 100 \tab 10 \tab 4 \tab 16 \tab 2 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 130 \tab 424 \tab 1160 \tab 407 \tab 1208 \tab 1057 \tab 2694 \tab 201 \tab 2119 \tab 408 \tab 783 \tab 657 \tab 683 \tab 682 \tab 1065 \tab 319 \tab 1619 \tab 1324 \tab 1811 \tab 3057 \tab 75 \tab 131 \tab 532 \tab 490 \tab 251 \tab 590 \tab 229 \tab 571 \tab 758 \tab 1433 \tab 394 \tab 20 \tab 920 \tab 377 \tab 522 \tab 216 \tab 392 \tab 258 \tab 1121 \tab 1022 \tab 117 \tab 2584 \tab 341 \tab 117 \tab 423 \tab 1227 \tab 3375 \tab 501 \tab 122 \tab 1022 \tab 369 \tab 337 \tab 330 \tab 780 \tab 687 \tab 488 \tab 799 \tab 359 \tab 565 \tab 497 \tab 115 \tab 72 \tab 121 \tab 317 \tab 525 \tab 195 \tab 382 \tab 304 \tab 70 \tab 584 \tab 169 \tab 162 \tab 624 \tab 172 \tab 330 \tab 392 \tab 1965 \tab 334 \tab 1187 \tab 407 \tab 344 \tab 330 \tab 586 \tab 1241 \tab 174 \tab 46 \tab 1046 \tab 313 \tab 8309 \tab 135 \tab 76 \tab 153 \tab 28 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
