% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_02.R
\name{prestg80}
\alias{prestg80}
\title{R's occupational prestige score(1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
prestg80
}
\details{
Question 3. Respondent's occupation, prestige, and industry
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 17 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 68 \tab 69 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 86 \tab dk, na, iap \tab 67 \tab 70 \tab iap \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 6 \tab 2 \tab 10 \tab 12 \tab 33 \tab 36 \tab 24 \tab 15 \tab 2 \tab 15 \tab 41 \tab 49 \tab 52 \tab 43 \tab 52 \tab 19 \tab 52 \tab 47 \tab 56 \tab 12 \tab 18 \tab 40 \tab 55 \tab 19 \tab 60 \tab 23 \tab 31 \tab 11 \tab 68 \tab 66 \tab 17 \tab 23 \tab 14 \tab 114 \tab 10 \tab 11 \tab 15 \tab 7 \tab 3 \tab 8 \tab 3 \tab 13 \tab 20 \tab 9 \tab 4 \tab 3 \tab 38 \tab 16 \tab 30 \tab 9 \tab 13 \tab 6 \tab 4 \tab 6 \tab 18 \tab 5 \tab 6 \tab 87 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 3 \tab 1 \tab 11 \tab 7 \tab 35 \tab 40 \tab 25 \tab 10 \tab 4 \tab 18 \tab 39 \tab 43 \tab 56 \tab 29 \tab 57 \tab 27 \tab 49 \tab 38 \tab 53 \tab 14 \tab 12 \tab 43 \tab 45 \tab 21 \tab 55 \tab 21 \tab 47 \tab 19 \tab 94 \tab 42 \tab 16 \tab 44 \tab 21 \tab 109 \tab 15 \tab 11 \tab 21 \tab 5 \tab 3 \tab 11 \tab 6 \tab 11 \tab 26 \tab 4 \tab 4 \tab 10 \tab 38 \tab 15 \tab 41 \tab 5 \tab 12 \tab 3 \tab 4 \tab 8 \tab 25 \tab 11 \tab 3 \tab 96 \tab 1 \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 5 \tab 2 \tab 5 \tab 5 \tab 30 \tab 33 \tab 20 \tab 11 \tab - \tab 14 \tab 38 \tab 45 \tab 60 \tab 28 \tab 50 \tab 21 \tab 46 \tab 42 \tab 72 \tab 5 \tab 12 \tab 42 \tab 50 \tab 13 \tab 46 \tab 19 \tab 28 \tab 22 \tab 76 \tab 54 \tab 13 \tab 36 \tab 11 \tab 89 \tab 21 \tab 7 \tab 12 \tab 2 \tab 1 \tab 10 \tab 1 \tab 6 \tab 20 \tab 10 \tab 3 \tab 3 \tab 51 \tab 20 \tab 38 \tab 4 \tab 11 \tab 3 \tab 1 \tab 4 \tab 19 \tab 6 \tab 4 \tab 72 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 4 \tab 3 \tab 5 \tab 7 \tab 38 \tab 21 \tab 14 \tab 13 \tab 5 \tab 14 \tab 58 \tab 55 \tab 50 \tab 44 \tab 47 \tab 25 \tab 24 \tab 30 \tab 80 \tab 11 \tab 11 \tab 42 \tab 54 \tab 20 \tab 56 \tab 21 \tab 38 \tab 26 \tab 72 \tab 59 \tab 24 \tab 35 \tab 26 \tab 103 \tab 14 \tab 4 \tab 15 \tab 8 \tab 5 \tab 9 \tab 8 \tab 12 \tab 28 \tab 13 \tab 6 \tab 5 \tab 51 \tab 21 \tab 37 \tab 6 \tab 7 \tab 3 \tab 1 \tab 6 \tab 11 \tab 7 \tab 4 \tab 99 \tab 2 \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 2 \tab 2 \tab 10 \tab 3 \tab 32 \tab 27 \tab 31 \tab 8 \tab 2 \tab 23 \tab 55 \tab 53 \tab 60 \tab 33 \tab 30 \tab 17 \tab 39 \tab 42 \tab 83 \tab 11 \tab 18 \tab 49 \tab 61 \tab 21 \tab 67 \tab 26 \tab 55 \tab 33 \tab 72 \tab 58 \tab 21 \tab 45 \tab 25 \tab 117 \tab 20 \tab 14 \tab 11 \tab 5 \tab 2 \tab 12 \tab 4 \tab 7 \tab 22 \tab 10 \tab 7 \tab 6 \tab 42 \tab 23 \tab 35 \tab 7 \tab 13 \tab 4 \tab 5 \tab 6 \tab 15 \tab 16 \tab 7 \tab 80 \tab - \tab 2 \tab - \tab 1606 \cr
   1994 \tab - \tab 16 \tab 10 \tab 20 \tab 12 \tab 72 \tab 51 \tab 37 \tab 20 \tab 4 \tab 28 \tab 87 \tab 114 \tab 110 \tab 67 \tab 78 \tab 71 \tab 67 \tab 77 \tab 135 \tab 19 \tab 26 \tab 63 \tab 101 \tab 36 \tab 144 \tab 46 \tab 97 \tab 38 \tab 131 \tab 111 \tab 33 \tab 88 \tab 33 \tab 178 \tab 49 \tab 29 \tab 20 \tab 7 \tab 2 \tab 31 \tab 10 \tab 21 \tab 56 \tab 36 \tab 12 \tab 17 \tab 104 \tab 43 \tab 68 \tab 14 \tab 26 \tab 7 \tab 2 \tab 15 \tab 41 \tab 14 \tab 9 \tab 139 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 18 \tab 6 \tab 19 \tab 7 \tab 77 \tab 44 \tab 54 \tab 22 \tab 9 \tab 36 \tab 76 \tab 89 \tab 106 \tab 62 \tab 106 \tab 52 \tab 82 \tab 84 \tab 128 \tab 11 \tab 24 \tab 89 \tab 82 \tab 36 \tab 110 \tab 51 \tab 88 \tab 37 \tab 98 \tab 132 \tab 29 \tab 40 \tab 34 \tab 244 \tab 38 \tab 15 \tab 28 \tab 10 \tab 6 \tab 27 \tab 8 \tab 25 \tab 45 \tab 34 \tab 8 \tab 13 \tab 88 \tab 35 \tab 78 \tab 9 \tab 42 \tab 9 \tab 6 \tab 11 \tab 38 \tab 10 \tab 15 \tab 123 \tab - \tab 1 \tab - \tab 2904 \cr
   1998 \tab - \tab 11 \tab 3 \tab 15 \tab 14 \tab 75 \tab 35 \tab 39 \tab 22 \tab - \tab 24 \tab 60 \tab 77 \tab 93 \tab 64 \tab 102 \tab 51 \tab 81 \tab 75 \tab 115 \tab 12 \tab 26 \tab 76 \tab 80 \tab 27 \tab 112 \tab 39 \tab 85 \tab 37 \tab 121 \tab 109 \tab 35 \tab 79 \tab 26 \tab 248 \tab 30 \tab 20 \tab 26 \tab 8 \tab 7 \tab 22 \tab 5 \tab 35 \tab 48 \tab 49 \tab 9 \tab 7 \tab 83 \tab 41 \tab 91 \tab 9 \tab 29 \tab 8 \tab 4 \tab 11 \tab 37 \tab 18 \tab 9 \tab 154 \tab 2 \tab 2 \tab - \tab 2832 \cr
   2000 \tab - \tab 16 \tab 3 \tab 26 \tab 24 \tab 76 \tab 48 \tab 35 \tab 18 \tab 4 \tab 21 \tab 73 \tab 79 \tab 91 \tab 64 \tab 87 \tab 69 \tab 63 \tab 76 \tab 118 \tab 10 \tab 15 \tab 66 \tab 82 \tab 24 \tab 118 \tab 28 \tab 74 \tab 39 \tab 80 \tab 99 \tab 29 \tab 77 \tab 25 \tab 254 \tab 39 \tab 18 \tab 13 \tab 17 \tab 3 \tab 29 \tab 9 \tab 35 \tab 57 \tab 20 \tab 14 \tab 15 \tab 101 \tab 52 \tab 89 \tab 6 \tab 32 \tab 2 \tab 4 \tab 12 \tab 63 \tab 11 \tab 7 \tab 157 \tab 1 \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 15 \tab 7 \tab 14 \tab 18 \tab 75 \tab 50 \tab 37 \tab 19 \tab 1 \tab 26 \tab 71 \tab 78 \tab 102 \tab 69 \tab 93 \tab 45 \tab 74 \tab 59 \tab 113 \tab 6 \tab 20 \tab 83 \tab 67 \tab 29 \tab 129 \tab 61 \tab 72 \tab 32 \tab 77 \tab 105 \tab 25 \tab 92 \tab 35 \tab 227 \tab 40 \tab 18 \tab 21 \tab 8 \tab 10 \tab 27 \tab 10 \tab 22 \tab 43 \tab 37 \tab 11 \tab 14 \tab 104 \tab 37 \tab 80 \tab 9 \tab 36 \tab 3 \tab 2 \tab 9 \tab 50 \tab 12 \tab 11 \tab 122 \tab 1 \tab 2 \tab - \tab 2765 \cr
   2004 \tab - \tab 21 \tab - \tab 15 \tab 18 \tab 57 \tab 37 \tab 31 \tab 26 \tab 6 \tab 21 \tab 73 \tab 77 \tab 89 \tab 60 \tab 91 \tab 46 \tab 58 \tab 67 \tab 96 \tab 7 \tab 13 \tab 67 \tab 70 \tab 32 \tab 117 \tab 58 \tab 89 \tab 32 \tab 70 \tab 104 \tab 31 \tab 90 \tab 29 \tab 268 \tab 45 \tab 28 \tab 18 \tab 11 \tab 6 \tab 26 \tab 7 \tab 28 \tab 42 \tab 36 \tab 5 \tab 17 \tab 120 \tab 47 \tab 85 \tab 16 \tab 42 \tab 9 \tab 3 \tab 12 \tab 58 \tab 16 \tab 13 \tab 153 \tab - \tab 3 \tab - \tab 2812 \cr
   2006 \tab - \tab 22 \tab 7 \tab 26 \tab 13 \tab 99 \tab 100 \tab 60 \tab 45 \tab 2 \tab 36 \tab 94 \tab 124 \tab 147 \tab 114 \tab 135 \tab 99 \tab 114 \tab 111 \tab 194 \tab 27 \tab 35 \tab 139 \tab 123 \tab 48 \tab 167 \tab 76 \tab 129 \tab 60 \tab 116 \tab 168 \tab 40 \tab 165 \tab 49 \tab 361 \tab 66 \tab 30 \tab 60 \tab 15 \tab 9 \tab 24 \tab 13 \tab 39 \tab 71 \tab 74 \tab 16 \tab 14 \tab 137 \tab 51 \tab 165 \tab 20 \tab 38 \tab 17 \tab 10 \tab 18 \tab 88 \tab 28 \tab 20 \tab 240 \tab - \tab 2 \tab - \tab 4510 \cr
   2008 \tab - \tab 13 \tab 5 \tab 14 \tab 5 \tab 45 \tab 37 \tab 25 \tab 13 \tab 1 \tab 16 \tab 53 \tab 55 \tab 78 \tab 66 \tab 66 \tab 61 \tab 50 \tab 45 \tab 86 \tab 14 \tab 16 \tab 62 \tab 44 \tab 20 \tab 74 \tab 29 \tab 31 \tab 26 \tab 56 \tab 79 \tab 29 \tab 46 \tab 26 \tab 187 \tab 25 \tab 14 \tab 16 \tab 11 \tab 1 \tab 21 \tab 6 \tab 29 \tab 33 \tab 26 \tab 13 \tab 11 \tab 51 \tab 30 \tab 59 \tab 7 \tab 23 \tab 8 \tab 6 \tab 6 \tab 35 \tab 12 \tab 5 \tab 102 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 8 \tab - \tab 10 \tab 7 \tab 63 \tab 31 \tab 24 \tab 19 \tab 3 \tab 19 \tab 58 \tab 66 \tab 62 \tab 65 \tab 79 \tab 54 \tab 49 \tab 35 \tab 75 \tab 8 \tab 17 \tab 42 \tab 48 \tab 26 \tab 70 \tab 35 \tab 55 \tab 29 \tab 56 \tab 81 \tab 22 \tab 45 \tab 10 \tab 188 \tab 21 \tab 13 \tab 20 \tab 13 \tab 4 \tab 23 \tab 5 \tab 20 \tab 28 \tab 18 \tab 7 \tab 9 \tab 73 \tab 24 \tab 59 \tab 13 \tab 17 \tab 10 \tab 4 \tab 7 \tab 30 \tab 10 \tab 2 \tab 154 \tab - \tab 1 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 21875 \tab 160 \tab 51 \tab 200 \tab 152 \tab 807 \tab 590 \tab 456 \tab 261 \tab 43 \tab 311 \tab 876 \tab 1004 \tab 1156 \tab 808 \tab 1073 \tab 657 \tab 848 \tab 828 \tab 1404 \tab 167 \tab 263 \tab 903 \tab 962 \tab 372 \tab 1325 \tab 533 \tab 919 \tab 441 \tab 1187 \tab 1267 \tab 364 \tab 905 \tab 364 \tab 2687 \tab 433 \tab 232 \tab 296 \tab 127 \tab 62 \tab 280 \tab 95 \tab 303 \tab 539 \tab 376 \tab 119 \tab 144 \tab 1081 \tab 455 \tab 955 \tab 134 \tab 341 \tab 92 \tab 56 \tab 131 \tab 528 \tab 176 \tab 115 \tab 15537 \tab 7 \tab 13 \tab 3544 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
