% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_46.R
\name{racchurh}
\alias{racchurh}
\title{racchurh}
\source{
General Social Survey https://gss.norc.org
}
\description{
Attend church with other race
}
\details{
Question 137. Do (blacks/Negroes/African-Americans)/whites attend the church that you, yourself, attend most often, or not?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab no church \tab yes \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 2 \tab 829 \tab 3 \tab 260 \tab 438 \tab 1532 \cr
   1980 \tab - \tab 2 \tab 719 \tab 4 \tab 223 \tab 520 \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 4 \tab 862 \tab 8 \tab 243 \tab 482 \tab 1599 \cr
   1984 \tab - \tab - \tab 692 \tab 14 \tab 204 \tab 563 \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 4 \tab 798 \tab 4 \tab 208 \tab 456 \tab 1470 \cr
   1987 \tab - \tab - \tab 881 \tab 30 \tab 196 \tab 712 \tab 1819 \cr
   1988 \tab 484 \tab 1 \tab 433 \tab 4 \tab 152 \tab 407 \tab 1481 \cr
   1989 \tab 502 \tab - \tab 498 \tab 5 \tab 192 \tab 340 \tab 1537 \cr
   1990 \tab 473 \tab 2 \tab 412 \tab 2 \tab 158 \tab 325 \tab 1372 \cr
   1991 \tab 500 \tab 1 \tab 471 \tab 17 \tab 138 \tab 390 \tab 1517 \cr
   1993 \tab 549 \tab 1 \tab 443 \tab 3 \tab 211 \tab 399 \tab 1606 \cr
   1994 \tab 2481 \tab 1 \tab 224 \tab 3 \tab 83 \tab 200 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 57513 \tab 18 \tab 7262 \tab 97 \tab 2268 \tab 5232 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} no church
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
