% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{racclos}
\alias{racclos}
\title{Any opp. race living close to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
racclos
}
\details{
Question 129a. Are there any (negro/black/African-American) families living close to you?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab don't know \tab Total \cr
   1972 \tab 1226 \tab 100 \tab 4 \tab 283 \tab - \tab 1613 \cr
   1973 \tab 980 \tab 148 \tab 7 \tab 369 \tab - \tab 1504 \cr
   1974 \tab 929 \tab 173 \tab 4 \tab 377 \tab 1 \tab 1484 \cr
   1975 \tab 1053 \tab 125 \tab 5 \tab 307 \tab - \tab 1490 \cr
   1976 \tab 916 \tab 158 \tab 7 \tab 417 \tab 1 \tab 1499 \cr
   1977 \tab 994 \tab 153 \tab 12 \tab 371 \tab - \tab 1530 \cr
   1978 \tab 785 \tab 202 \tab 2 \tab 543 \tab - \tab 1532 \cr
   1980 \tab 791 \tab 178 \tab 3 \tab 495 \tab 1 \tab 1468 \cr
   1982 \tab 918 \tab 212 \tab 8 \tab 721 \tab 1 \tab 1860 \cr
   1983 \tab 845 \tab 227 \tab 6 \tab 521 \tab - \tab 1599 \cr
   1984 \tab 719 \tab 164 \tab 17 \tab 573 \tab - \tab 1473 \cr
   1985 \tab 793 \tab 183 \tab 16 \tab 542 \tab - \tab 1534 \cr
   1986 \tab 767 \tab 185 \tab 8 \tab 506 \tab 4 \tab 1470 \cr
   1987 \tab - \tab 814 \tab 51 \tab 954 \tab - \tab 1819 \cr
   1988 \tab 686 \tab 155 \tab 16 \tab 621 \tab 3 \tab 1481 \cr
   1989 \tab 741 \tab 217 \tab 19 \tab 560 \tab - \tab 1537 \cr
   1990 \tab 625 \tab 171 \tab 15 \tab 559 \tab 2 \tab 1372 \cr
   1991 \tab 714 \tab 176 \tab 25 \tab 600 \tab 2 \tab 1517 \cr
   1993 \tab 641 \tab 235 \tab 25 \tab 699 \tab 6 \tab 1606 \cr
   1994 \tab 2541 \tab 100 \tab 16 \tab 333 \tab 2 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 57674 \tab 4076 \tab 266 \tab 10351 \tab 23 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
