% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_81.R
\name{racdif2}
\alias{racdif2}
\title{Differences due to in-born learning ability}
\source{
General Social Survey https://gss.norc.org
}
\description{
racdif2
}
\details{
Question 266. On the average (negroes/blacks/African-Americans) have worse jobs, income, and housing than white people. Do you think these differences are: b. Because most (negroes/blacks/African-Americans) have less in-born ability to learn?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 176 \tab 66 \tab 947 \tab 7 \tab 334 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 62 \tab 1157 \tab 5 \tab 310 \tab - \tab 1534 \cr
   1986 \tab - \tab 52 \tab 1123 \tab 3 \tab 292 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 40 \tab 761 \tab 7 \tab 180 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 45 \tab 774 \tab 6 \tab 181 \tab - \tab 1537 \cr
   1990 \tab - \tab 43 \tab 1081 \tab 8 \tab 240 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 44 \tab 831 \tab 6 \tab 143 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 37 \tab 903 \tab 5 \tab 135 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 79 \tab 1634 \tab 11 \tab 253 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 65 \tab 1697 \tab 5 \tab 193 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 77 \tab 1610 \tab 6 \tab 178 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 68 \tab 1573 \tab 9 \tab 227 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 21 \tab 781 \tab - \tab 106 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 19 \tab 804 \tab 3 \tab 80 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 55 \tab 1759 \tab 5 \tab 173 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 28 \tab 1158 \tab 3 \tab 140 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 30 \tab 1248 \tab 8 \tab 144 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 28 \tab 1146 \tab 3 \tab 125 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 31 \tab 1496 \tab 6 \tab 142 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 28 \tab 1703 \tab 3 \tab 154 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 24 \tab 1402 \tab 4 \tab 129 \tab - \tab 2348 \cr
   2021 \tab 1315 \tab 11 \tab 2548 \tab 1 \tab 141 \tab 16 \tab 4032 \cr
   2022 \tab 2373 \tab 18 \tab 1063 \tab - \tab 86 \tab 4 \tab 3544 \cr
   Total \tab 38000 \tab 971 \tab 29199 \tab 114 \tab 4086 \tab 20 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
