% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{racdis}
\alias{racdis}
\title{racdis}
\source{
General Social Survey https://gss.norc.org
}
\description{
How far are closest opp. race
}
\details{
Question 129b. How many blocks (or miles) away do they live (the Negro/black/African American] families who live closest to you) live?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab 1-3 blks away \tab 4-8 blks away \tab iap \tab no answer \tab over 8 blks \tab same block \tab don't know \tab Total \cr
   1972 \tab 150 \tab 54 \tab 1226 \tab 7 \tab 22 \tab 154 \tab - \tab 1613 \cr
   1973 \tab 208 \tab 80 \tab 980 \tab 5 \tab 31 \tab 199 \tab 1 \tab 1504 \cr
   1974 \tab 190 \tab 89 \tab 929 \tab 8 \tab 36 \tab 231 \tab 1 \tab 1484 \cr
   1975 \tab 151 \tab 64 \tab 1053 \tab 4 \tab 37 \tab 178 \tab 3 \tab 1490 \cr
   1976 \tab 185 \tab 97 \tab 916 \tab 8 \tab 50 \tab 243 \tab - \tab 1499 \cr
   1977 \tab 178 \tab 98 \tab 994 \tab 9 \tab 42 \tab 207 \tab 2 \tab 1530 \cr
   1978 \tab 266 \tab 88 \tab 785 \tab 10 \tab 48 \tab 335 \tab - \tab 1532 \cr
   1980 \tab 223 \tab 103 \tab 791 \tab 8 \tab 26 \tab 311 \tab 6 \tab 1468 \cr
   1982 \tab 289 \tab 106 \tab 918 \tab 19 \tab 46 \tab 479 \tab 3 \tab 1860 \cr
   1983 \tab 234 \tab 93 \tab 845 \tab 8 \tab 62 \tab 351 \tab 6 \tab 1599 \cr
   1984 \tab 224 \tab 62 \tab 719 \tab 24 \tab 34 \tab 409 \tab 1 \tab 1473 \cr
   1985 \tab 205 \tab 99 \tab 793 \tab 17 \tab 44 \tab 375 \tab 1 \tab 1534 \cr
   1986 \tab 211 \tab 70 \tab 767 \tab 9 \tab 45 \tab 362 \tab 6 \tab 1470 \cr
   1987 \tab 345 \tab 241 \tab - \tab 88 \tab 550 \tab 595 \tab - \tab 1819 \cr
   1988 \tab 248 \tab 65 \tab 686 \tab 17 \tab 32 \tab 429 \tab 4 \tab 1481 \cr
   1989 \tab 263 \tab 101 \tab 741 \tab 21 \tab 50 \tab 360 \tab 1 \tab 1537 \cr
   1990 \tab 239 \tab 85 \tab 625 \tab 9 \tab 43 \tab 368 \tab 3 \tab 1372 \cr
   1991 \tab 234 \tab 68 \tab 714 \tab 27 \tab 47 \tab 422 \tab 5 \tab 1517 \cr
   1993 \tab 277 \tab 101 \tab 641 \tab 27 \tab 65 \tab 482 \tab 13 \tab 1606 \cr
   1994 \tab 125 \tab 35 \tab 2541 \tab 16 \tab 20 \tab 252 \tab 3 \tab 2992 \cr
   1996 \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 4445 \tab 1799 \tab 57674 \tab 341 \tab 1330 \tab 6742 \tab 59 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} same block
\item \code{2} 1-3 blks away
\item \code{3} 4-8 blks away
\item \code{4} over 8 blks
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
