% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{racfew}
\alias{racfew}
\title{Children to school with opposite race}
\source{
General Social Survey https://gss.norc.org
}
\description{
racfew
}
\details{
Question 133a. Would you yourself have any objection to sending your children to a school wehre a few of the children are (whites/(negroes/blacks/African-Americans))?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab don't know \tab iap \tab no \tab no answer \tab yes, object \tab Total \cr
   1972 \tab 13 \tab 261 \tab 1242 \tab 1 \tab 96 \tab 1613 \cr
   1973 \tab - \tab 1504 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 7 \tab 173 \tab 1221 \tab 16 \tab 67 \tab 1484 \cr
   1975 \tab 22 \tab 163 \tab 1214 \tab 1 \tab 90 \tab 1490 \cr
   1976 \tab - \tab 1499 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 5 \tab 176 \tab 1247 \tab 4 \tab 98 \tab 1530 \cr
   1978 \tab 15 \tab - \tab 1441 \tab 6 \tab 70 \tab 1532 \cr
   1980 \tab - \tab 1468 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 17 \tab - \tab 1735 \tab 4 \tab 104 \tab 1860 \cr
   1983 \tab 15 \tab - \tab 1503 \tab 3 \tab 78 \tab 1599 \cr
   1984 \tab - \tab 1473 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 8 \tab - \tab 1454 \tab 8 \tab 64 \tab 1534 \cr
   1986 \tab 11 \tab - \tab 1386 \tab 3 \tab 70 \tab 1470 \cr
   1987 \tab - \tab 1819 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 13 \tab 493 \tab 932 \tab 2 \tab 41 \tab 1481 \cr
   1989 \tab 18 \tab 531 \tab 942 \tab 4 \tab 42 \tab 1537 \cr
   1990 \tab 11 \tab - \tab 1314 \tab 5 \tab 42 \tab 1372 \cr
   1991 \tab 10 \tab 493 \tab 973 \tab 8 \tab 33 \tab 1517 \cr
   1993 \tab 5 \tab 526 \tab 1026 \tab 2 \tab 47 \tab 1606 \cr
   1994 \tab 18 \tab 1015 \tab 1875 \tab 13 \tab 71 \tab 2992 \cr
   1996 \tab 5 \tab 1913 \tab 951 \tab 6 \tab 29 \tab 2904 \cr
   1998 \tab - \tab 2832 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 2817 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 2765 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 2812 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 4510 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 2023 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 2044 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1974 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2538 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2867 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 2348 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 4032 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab 3544 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 193 \tab 50613 \tab 20456 \tab 86 \tab 1042 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes, object
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
