% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{racinteg}
\alias{racinteg}
\title{Will neighborhood become all black}
\source{
General Social Survey https://gss.norc.org
}
\description{
racinteg
}
\details{
Question 129c. Do you think this neighborhood will become all (negro/black/African-American) in the next few years, or will it remain integrated?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab all black \tab don't know \tab iap \tab no answer \tab stay integrated \tab mixed,not wh-bl \tab Total \cr
   1972 \tab 38 \tab 28 \tab 1226 \tab 9 \tab 312 \tab - \tab 1613 \cr
   1973 \tab 26 \tab 23 \tab 980 \tab 8 \tab 466 \tab 1 \tab 1504 \cr
   1974 \tab 26 \tab 29 \tab 929 \tab 6 \tab 493 \tab 1 \tab 1484 \cr
   1975 \tab 27 \tab 23 \tab 1053 \tab 6 \tab 381 \tab - \tab 1490 \cr
   1976 \tab 37 \tab 17 \tab 916 \tab 8 \tab 521 \tab - \tab 1499 \cr
   1977 \tab 24 \tab 17 \tab 994 \tab 12 \tab 483 \tab - \tab 1530 \cr
   1978 \tab 47 \tab 19 \tab 785 \tab 21 \tab 660 \tab - \tab 1532 \cr
   1980 \tab 42 \tab 26 \tab 791 \tab 12 \tab 597 \tab - \tab 1468 \cr
   1982 \tab 103 \tab 46 \tab 918 \tab 18 \tab 775 \tab - \tab 1860 \cr
   1983 \tab 42 \tab 35 \tab 845 \tab 10 \tab 667 \tab - \tab 1599 \cr
   1984 \tab 41 \tab 13 \tab 719 \tab 27 \tab 673 \tab - \tab 1473 \cr
   1985 \tab 33 \tab 24 \tab 793 \tab 20 \tab 664 \tab - \tab 1534 \cr
   1986 \tab 43 \tab 19 \tab 767 \tab 5 \tab 636 \tab - \tab 1470 \cr
   1987 \tab 89 \tab 39 \tab 832 \tab 18 \tab 841 \tab - \tab 1819 \cr
   1988 \tab 44 \tab 19 \tab 686 \tab 19 \tab 713 \tab - \tab 1481 \cr
   1989 \tab 34 \tab 26 \tab 741 \tab 20 \tab 716 \tab - \tab 1537 \cr
   1990 \tab 45 \tab 27 \tab 625 \tab 14 \tab 661 \tab - \tab 1372 \cr
   1991 \tab 43 \tab 21 \tab 714 \tab 21 \tab 718 \tab - \tab 1517 \cr
   1993 \tab 40 \tab 26 \tab 641 \tab 16 \tab 883 \tab - \tab 1606 \cr
   1994 \tab 28 \tab 20 \tab 2541 \tab 8 \tab 395 \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 852 \tab 497 \tab 58506 \tab 278 \tab 12255 \tab 2 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} all black
\item \code{2} stay integrated
\item \code{3} mixed,not wh-bl
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
