% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{raclive}
\alias{raclive}
\title{Any opp. race in neighborhood}
\source{
General Social Survey https://gss.norc.org
}
\description{
raclive
}
\details{
Question 129. Are there any (negroes/blacks/African-Americans) living in this neighborhood now?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab don't know \tab iap \tab no \tab no answer \tab yes \tab skipped on web \tab Total \cr
   1972 \tab 57 \tab 261 \tab 908 \tab 1 \tab 386 \tab - \tab 1613 \cr
   1973 \tab 51 \tab 183 \tab 745 \tab 1 \tab 524 \tab - \tab 1504 \cr
   1974 \tab 45 \tab 173 \tab 711 \tab 1 \tab 554 \tab - \tab 1484 \cr
   1975 \tab 56 \tab 163 \tab 834 \tab 1 \tab 436 \tab - \tab 1490 \cr
   1976 \tab 50 \tab 129 \tab 737 \tab 3 \tab 580 \tab - \tab 1499 \cr
   1977 \tab 41 \tab 176 \tab 777 \tab 4 \tab 532 \tab - \tab 1530 \cr
   1978 \tab 54 \tab - \tab 731 \tab 2 \tab 745 \tab - \tab 1532 \cr
   1980 \tab 54 \tab - \tab 737 \tab 2 \tab 675 \tab - \tab 1468 \cr
   1982 \tab 66 \tab - \tab 852 \tab 4 \tab 938 \tab - \tab 1860 \cr
   1983 \tab 49 \tab - \tab 796 \tab 3 \tab 751 \tab - \tab 1599 \cr
   1984 \tab 59 \tab - \tab 660 \tab 8 \tab 746 \tab - \tab 1473 \cr
   1985 \tab 67 \tab - \tab 726 \tab 7 \tab 734 \tab - \tab 1534 \cr
   1986 \tab 40 \tab - \tab 727 \tab 2 \tab 701 \tab - \tab 1470 \cr
   1987 \tab 60 \tab - \tab 772 \tab 6 \tab 981 \tab - \tab 1819 \cr
   1988 \tab 67 \tab - \tab 619 \tab 5 \tab 790 \tab - \tab 1481 \cr
   1989 \tab 75 \tab - \tab 666 \tab 10 \tab 786 \tab - \tab 1537 \cr
   1990 \tab 75 \tab - \tab 550 \tab 4 \tab 743 \tab - \tab 1372 \cr
   1991 \tab 66 \tab - \tab 648 \tab 15 \tab 788 \tab - \tab 1517 \cr
   1993 \tab 79 \tab - \tab 562 \tab 10 \tab 955 \tab - \tab 1606 \cr
   1994 \tab 143 \tab - \tab 1014 \tab 15 \tab 1820 \tab - \tab 2992 \cr
   1996 \tab 158 \tab - \tab 949 \tab 26 \tab 1771 \tab - \tab 2904 \cr
   1998 \tab 141 \tab - \tab 887 \tab 55 \tab 1749 \tab - \tab 2832 \cr
   2000 \tab 150 \tab - \tab 831 \tab 45 \tab 1791 \tab - \tab 2817 \cr
   2002 \tab 49 \tab 1395 \tab 365 \tab 23 \tab 933 \tab - \tab 2765 \cr
   2004 \tab 55 \tab 1472 \tab 431 \tab 1 \tab 853 \tab - \tab 2812 \cr
   2006 \tab 91 \tab 1518 \tab 927 \tab 1 \tab 1973 \tab - \tab 4510 \cr
   2008 \tab 77 \tab - \tab 580 \tab - \tab 1366 \tab - \tab 2023 \cr
   2010 \tab 85 \tab - \tab 596 \tab 2 \tab 1361 \tab - \tab 2044 \cr
   2012 \tab 82 \tab - \tab 517 \tab 1 \tab 1374 \tab - \tab 1974 \cr
   2014 \tab 118 \tab - \tab 587 \tab 2 \tab 1831 \tab - \tab 2538 \cr
   2016 \tab 122 \tab - \tab 648 \tab 2 \tab 2095 \tab - \tab 2867 \cr
   2018 \tab 127 \tab - \tab 477 \tab 6 \tab 1738 \tab - \tab 2348 \cr
   2021 \tab 440 \tab - \tab 632 \tab - \tab 2955 \tab 5 \tab 4032 \cr
   2022 \tab 257 \tab - \tab 665 \tab - \tab 2620 \tab 2 \tab 3544 \cr
   Total \tab 3206 \tab 5470 \tab 23864 \tab 268 \tab 39575 \tab 7 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
