% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_44.R
\name{racmar}
\alias{racmar}
\title{Favor law against racial intermarriage}
\source{
General Social Survey https://gss.norc.org
}
\description{
racmar
}
\details{
Question 125a. Do you think there should be laws against marriages between (Negroes/Blacks/African-Americans) and whites?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab don't know \tab iap \tab no \tab yes \tab no answer \tab Total \cr
   1972 \tab 43 \tab 261 \tab 795 \tab 514 \tab - \tab 1613 \cr
   1973 \tab 20 \tab 183 \tab 801 \tab 488 \tab 12 \tab 1504 \cr
   1974 \tab 29 \tab 173 \tab 838 \tab 442 \tab 2 \tab 1484 \cr
   1975 \tab 29 \tab 163 \tab 793 \tab 499 \tab 6 \tab 1490 \cr
   1976 \tab 33 \tab 129 \tab 897 \tab 433 \tab 7 \tab 1499 \cr
   1977 \tab 21 \tab 176 \tab 952 \tab 375 \tab 6 \tab 1530 \cr
   1978 \tab - \tab 1532 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 39 \tab - \tab 1000 \tab 427 \tab 2 \tab 1468 \cr
   1982 \tab 53 \tab - \tab 1339 \tab 463 \tab 5 \tab 1860 \cr
   1983 \tab - \tab 1599 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 37 \tab - \tab 1037 \tab 346 \tab 53 \tab 1473 \cr
   1985 \tab 38 \tab - \tab 1099 \tab 387 \tab 10 \tab 1534 \cr
   1986 \tab - \tab 1470 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 54 \tab - \tab 1390 \tab 370 \tab 5 \tab 1819 \cr
   1988 \tab 29 \tab 504 \tab 732 \tab 213 \tab 3 \tab 1481 \cr
   1989 \tab 27 \tab 504 \tab 790 \tab 214 \tab 2 \tab 1537 \cr
   1990 \tab 26 \tab 455 \tab 719 \tab 171 \tab 1 \tab 1372 \cr
   1991 \tab 32 \tab 524 \tab 786 \tab 168 \tab 7 \tab 1517 \cr
   1993 \tab 35 \tab 531 \tab 859 \tab 178 \tab 3 \tab 1606 \cr
   1994 \tab 46 \tab 996 \tab 1678 \tab 268 \tab 4 \tab 2992 \cr
   1996 \tab 41 \tab 981 \tab 1672 \tab 209 \tab 1 \tab 2904 \cr
   1998 \tab 50 \tab 950 \tab 1625 \tab 204 \tab 3 \tab 2832 \cr
   2000 \tab 47 \tab 1067 \tab 1525 \tab 172 \tab 6 \tab 2817 \cr
   2002 \tab 7 \tab 1841 \tab 828 \tab 88 \tab 1 \tab 2765 \cr
   2004 \tab - \tab 2812 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 4510 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 2023 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 2044 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1974 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2538 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2867 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 2348 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 4032 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab 3544 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 736 \tab 42731 \tab 22155 \tab 6629 \tab 139 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
