% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_46.R
\name{racmost}
\alias{racmost}
\title{Children to school mostly opposite race}
\source{
General Social Survey https://gss.norc.org
}
\description{
racmost
}
\details{
Question 133c. Where more than half of the children are (whites/(negroes/blacks/African-Americans))?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab don't know \tab iap \tab no \tab no answer \tab yes, object \tab Total \cr
   1972 \tab 65 \tab 573 \tab 562 \tab 12 \tab 401 \tab 1613 \cr
   1973 \tab - \tab 1504 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 61 \tab 541 \tab 454 \tab 21 \tab 407 \tab 1484 \cr
   1975 \tab 68 \tab 524 \tab 490 \tab 11 \tab 397 \tab 1490 \cr
   1976 \tab - \tab 1499 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 41 \tab 506 \tab 517 \tab 13 \tab 453 \tab 1530 \cr
   1978 \tab 59 \tab 329 \tab 703 \tab 14 \tab 427 \tab 1532 \cr
   1980 \tab - \tab 1468 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 73 \tab 360 \tab 939 \tab 18 \tab 470 \tab 1860 \cr
   1983 \tab 64 \tab 345 \tab 647 \tab 15 \tab 528 \tab 1599 \cr
   1984 \tab - \tab 1473 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 55 \tab 301 \tab 678 \tab 14 \tab 486 \tab 1534 \cr
   1986 \tab 41 \tab 316 \tab 624 \tab 10 \tab 479 \tab 1470 \cr
   1987 \tab - \tab 1819 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 44 \tab 682 \tab 460 \tab 5 \tab 290 \tab 1481 \cr
   1989 \tab 48 \tab 736 \tab 430 \tab 10 \tab 313 \tab 1537 \cr
   1990 \tab 74 \tab 255 \tab 590 \tab 15 \tab 438 \tab 1372 \cr
   1991 \tab 48 \tab 672 \tab 501 \tab 12 \tab 284 \tab 1517 \cr
   1993 \tab 53 \tab 721 \tab 513 \tab 8 \tab 311 \tab 1606 \cr
   1994 \tab 89 \tab 1318 \tab 994 \tab 20 \tab 571 \tab 2992 \cr
   1996 \tab 47 \tab 2082 \tab 516 \tab 25 \tab 234 \tab 2904 \cr
   1998 \tab - \tab 2832 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 2817 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 2765 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 2812 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 4510 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 2023 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 2044 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1974 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2538 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2867 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 2348 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 4032 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab 3544 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 930 \tab 55130 \tab 9618 \tab 223 \tab 6489 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes, object
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
