% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{racopen}
\alias{racopen}
\title{racopen}
\source{
General Social Survey https://gss.norc.org
}
\description{
Vote on open housing law
}
\details{
Question 128. Suppose there is a community-wide vote on the general housing issue. There are two possible laws to vote on: a. One law says that a homeowner can decide for himself whom to sell his house to, even if he prefers not to sell to (negroes/blacks/African-Americans). b. The second law says that a homeowner cannot refuse to sell to someone because of their race or color. Which law would you vote for?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab don't know \tab first law \tab neither \tab no answer \tab second law \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 183 \tab 18 \tab 826 \tab 22 \tab 7 \tab 448 \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 163 \tab 21 \tab 844 \tab 10 \tab 2 \tab 450 \tab - \tab 1490 \cr
   1976 \tab 129 \tab 17 \tab 851 \tab 24 \tab 7 \tab 471 \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 30 \tab 863 \tab 19 \tab 3 \tab 617 \tab - \tab 1532 \cr
   1980 \tab - \tab 25 \tab 806 \tab 23 \tab 2 \tab 612 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 35 \tab 779 \tab 30 \tab 7 \tab 748 \tab - \tab 1599 \cr
   1984 \tab - \tab 22 \tab 662 \tab 26 \tab 9 \tab 754 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 18 \tab 684 \tab 20 \tab 4 \tab 744 \tab - \tab 1470 \cr
   1987 \tab - \tab 31 \tab 697 \tab 31 \tab 14 \tab 1046 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 19 \tab 388 \tab 25 \tab 3 \tab 562 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 16 \tab 380 \tab 21 \tab 7 \tab 611 \tab - \tab 1537 \cr
   1990 \tab - \tab 39 \tab 549 \tab 22 \tab 5 \tab 757 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 17 \tab 352 \tab 29 \tab 14 \tab 605 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 19 \tab 314 \tab 40 \tab 4 \tab 680 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 36 \tab 652 \tab 58 \tab 14 \tab 1251 \tab - \tab 2992 \cr
   1996 \tab 1926 \tab 15 \tab 281 \tab 20 \tab 11 \tab 651 \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 13 \tab 289 \tab - \tab 4 \tab 592 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 8 \tab 542 \tab 94 \tab 6 \tab 1353 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 11 \tab 342 \tab 59 \tab 3 \tab 937 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 4 \tab 307 \tab 52 \tab 2 \tab 916 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 7 \tab 334 \tab 33 \tab 4 \tab 930 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 6 \tab 411 \tab 42 \tab 3 \tab 1252 \tab - \tab 2538 \cr
   2016 \tab 978 \tab 13 \tab 365 \tab 58 \tab 7 \tab 1446 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 14 \tab 304 \tab 27 \tab 4 \tab 1225 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 2943 \tab 5 \tab 108 \tab - \tab 8 \tab 480 \tab - \tab 3544 \cr
   Total \tab 33892 \tab 459 \tab 12930 \tab 785 \tab 154 \tab 20138 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} first law
\item \code{2} second law
\item \code{3} neither
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
