% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_46.R
\name{racpres}
\alias{racpres}
\title{racpres}
\source{
General Social Survey https://gss.norc.org
}
\description{
Would vote for black president
}
\details{
Question 135. If your party nominated a (negro/black/African-American) for President, would you vote for him if he were qualified for the job?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab don't know \tab iap \tab no \tab no answer \tab yes \tab Total \cr
   1972 \tab 84 \tab 261 \tab 332 \tab 3 \tab 933 \tab 1613 \cr
   1973 \tab - \tab 1504 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 56 \tab - \tab 242 \tab 5 \tab 1181 \tab 1484 \cr
   1975 \tab 80 \tab 163 \tab 227 \tab - \tab 1020 \tab 1490 \cr
   1976 \tab - \tab 1499 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 49 \tab 176 \tab 291 \tab 7 \tab 1007 \tab 1530 \cr
   1978 \tab 57 \tab - \tab 224 \tab 15 \tab 1236 \tab 1532 \cr
   1980 \tab - \tab 1468 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 84 \tab - \tab 181 \tab 7 \tab 1588 \tab 1860 \cr
   1983 \tab 66 \tab - \tab 217 \tab 7 \tab 1309 \tab 1599 \cr
   1984 \tab - \tab 1473 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 65 \tab - \tab 221 \tab 10 \tab 1238 \tab 1534 \cr
   1986 \tab 39 \tab - \tab 180 \tab 4 \tab 1247 \tab 1470 \cr
   1987 \tab - \tab 1819 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 46 \tab 493 \tab 168 \tab 5 \tab 769 \tab 1481 \cr
   1989 \tab 49 \tab 531 \tab 163 \tab 10 \tab 784 \tab 1537 \cr
   1990 \tab 42 \tab 444 \tab 105 \tab 8 \tab 773 \tab 1372 \cr
   1991 \tab 37 \tab 493 \tab 100 \tab 2 \tab 885 \tab 1517 \cr
   1993 \tab 55 \tab 526 \tab 113 \tab 3 \tab 909 \tab 1606 \cr
   1994 \tab 65 \tab 1015 \tab 174 \tab 11 \tab 1727 \tab 2992 \cr
   1996 \tab 29 \tab 1913 \tab 67 \tab 4 \tab 891 \tab 2904 \cr
   1998 \tab - \tab 2832 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 2817 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 2765 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 2812 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 4510 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 19 \tab 694 \tab 80 \tab 6 \tab 1224 \tab 2023 \cr
   2010 \tab 17 \tab 614 \tab 55 \tab 7 \tab 1351 \tab 2044 \cr
   2012 \tab - \tab 1974 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2538 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2867 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 2348 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 4032 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab 3544 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 939 \tab 48125 \tab 3140 \tab 114 \tab 20072 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
