% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{racschol}
\alias{racschol}
\title{racschol}
\source{
General Social Survey https://gss.norc.org
}
\description{
Should whites & blacks go to same school
}
\details{
Question 132. Do you think white students and (negro/black) students should go to teh same schools or to separate schools?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab don't know \tab no answer \tab same schools \tab separate schools \tab iap \tab Total \cr
   1972 \tab 32 \tab 7 \tab 1381 \tab 193 \tab - \tab 1613 \cr
   1973 \tab - \tab - \tab - \tab - \tab 1504 \tab 1504 \cr
   1974 \tab - \tab - \tab - \tab - \tab 1484 \tab 1484 \cr
   1975 \tab - \tab - \tab - \tab - \tab 1490 \tab 1490 \cr
   1976 \tab 39 \tab 3 \tab 1250 \tab 207 \tab - \tab 1499 \cr
   1977 \tab 25 \tab 10 \tab 1298 \tab 197 \tab - \tab 1530 \cr
   1978 \tab - \tab - \tab - \tab - \tab 1532 \tab 1532 \cr
   1980 \tab 29 \tab 6 \tab 1273 \tab 160 \tab - \tab 1468 \cr
   1982 \tab 43 \tab 6 \tab 1670 \tab 141 \tab - \tab 1860 \cr
   1983 \tab - \tab - \tab - \tab - \tab 1599 \tab 1599 \cr
   1984 \tab 26 \tab 10 \tab 1327 \tab 110 \tab - \tab 1473 \cr
   1985 \tab 8 \tab 5 \tab 686 \tab 52 \tab 783 \tab 1534 \cr
   1986 \tab - \tab - \tab - \tab - \tab 1470 \tab 1470 \cr
   1987 \tab - \tab - \tab - \tab - \tab 1819 \tab 1819 \cr
   1988 \tab - \tab - \tab - \tab - \tab 1481 \tab 1481 \cr
   1989 \tab - \tab - \tab - \tab - \tab 1537 \tab 1537 \cr
   1990 \tab - \tab - \tab - \tab - \tab 1372 \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab 1517 \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab 1606 \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab 2992 \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab 2904 \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab 2832 \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab 2817 \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab 2765 \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab 2812 \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab 4510 \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab 2023 \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab 2044 \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab 1974 \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab 2538 \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab 2867 \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab 2348 \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 202 \tab 47 \tab 8885 \tab 1060 \tab 62196 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} same schools
\item \code{2} separate schools
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
