% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_44.R
\name{racseg}
\alias{racseg}
\title{racseg}
\source{
General Social Survey https://gss.norc.org
}
\description{
Whites have right to seg. neighborhood
}
\details{
Question 127. Here are some opinions other people have expressed in connection with (negro/black)-white relations. Do you agree or disagree with each statement: b. White people have a right to keep (negroes/blacks/African-Americans) out of their neighborhoods if they want to, and (negroes/blacks/African-Americans) should respect that right.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab agree slightly \tab agree strongly \tab disagree slightly \tab disagree strongly \tab don't know \tab iap \tab no answer \tab Total \cr
   1972 \tab 224 \tab 279 \tab 307 \tab 443 \tab 96 \tab 261 \tab 3 \tab 1613 \cr
   1973 \tab - \tab - \tab - \tab - \tab - \tab 1504 \tab - \tab 1504 \cr
   1974 \tab - \tab - \tab - \tab - \tab - \tab 1484 \tab - \tab 1484 \cr
   1975 \tab - \tab - \tab - \tab - \tab - \tab 1490 \tab - \tab 1490 \cr
   1976 \tab 237 \tab 288 \tab 346 \tab 466 \tab 26 \tab 129 \tab 7 \tab 1499 \cr
   1977 \tab 266 \tab 289 \tab 381 \tab 373 \tab 40 \tab 176 \tab 5 \tab 1530 \cr
   1978 \tab - \tab - \tab - \tab - \tab - \tab 1532 \tab - \tab 1532 \cr
   1980 \tab 220 \tab 221 \tab 397 \tab 576 \tab 50 \tab - \tab 4 \tab 1468 \cr
   1982 \tab 222 \tab 220 \tab 452 \tab 923 \tab 39 \tab - \tab 4 \tab 1860 \cr
   1983 \tab - \tab - \tab - \tab - \tab - \tab 1599 \tab - \tab 1599 \cr
   1984 \tab 211 \tab 151 \tab 368 \tab 705 \tab 31 \tab - \tab 7 \tab 1473 \cr
   1985 \tab 224 \tab 151 \tab 432 \tab 680 \tab 38 \tab - \tab 9 \tab 1534 \cr
   1986 \tab - \tab - \tab - \tab - \tab - \tab 1470 \tab - \tab 1470 \cr
   1987 \tab 217 \tab 149 \tab 401 \tab 1009 \tab 38 \tab - \tab 5 \tab 1819 \cr
   1988 \tab 138 \tab 72 \tab 230 \tab 510 \tab 26 \tab 504 \tab 1 \tab 1481 \cr
   1989 \tab 136 \tab 78 \tab 234 \tab 553 \tab 28 \tab 504 \tab 4 \tab 1537 \cr
   1990 \tab 124 \tab 74 \tab 218 \tab 471 \tab 27 \tab 455 \tab 3 \tab 1372 \cr
   1991 \tab 105 \tab 66 \tab 221 \tab 574 \tab 21 \tab 524 \tab 6 \tab 1517 \cr
   1993 \tab 109 \tab 42 \tab 218 \tab 681 \tab 23 \tab 531 \tab 2 \tab 1606 \cr
   1994 \tab 197 \tab 102 \tab 472 \tab 1176 \tab 44 \tab 996 \tab 5 \tab 2992 \cr
   1996 \tab 59 \tab 50 \tab 177 \tab 636 \tab 27 \tab 1953 \tab 2 \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab 3544 \cr
   Total \tab 2689 \tab 2232 \tab 4854 \tab 9776 \tab 554 \tab 52218 \tab 67 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree slightly
\item \code{3} disagree slightly
\item \code{4} disagree strongly
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
