% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_567.R
\name{realinc}
\alias{realinc}
\title{realinc}
\source{
General Social Survey https://gss.norc.org
}
\description{
Family income in constant $
}
\details{
Question 1577. Family income on 1972-2006 surveys in constant dollars (base = 1986)
}
\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
