% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{reg16}
\alias{reg16}
\title{Region of residence, age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
reg16
}
\details{
Question 26. In what state or foreign country were you living when you were 16 years old?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab east north central \tab east south atlantic \tab foreign \tab middle atlantic \tab mountain \tab new england \tab pacific \tab south atlantic \tab west north central \tab west south central \tab iap \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 282 \tab 119 \tab 66 \tab 300 \tab 43 \tab 88 \tab 132 \tab 196 \tab 178 \tab 209 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 298 \tab 110 \tab 58 \tab 258 \tab 42 \tab 59 \tab 129 \tab 288 \tab 129 \tab 133 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 322 \tab 103 \tab 54 \tab 253 \tab 42 \tab 59 \tab 120 \tab 259 \tab 143 \tab 129 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 311 \tab 120 \tab 50 \tab 260 \tab 57 \tab 55 \tab 113 \tab 272 \tab 125 \tab 127 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 315 \tab 111 \tab 66 \tab 263 \tab 51 \tab 64 \tab 114 \tab 250 \tab 143 \tab 122 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 357 \tab 100 \tab 73 \tab 243 \tab 47 \tab 69 \tab 130 \tab 279 \tab 112 \tab 120 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 327 \tab 107 \tab 61 \tab 269 \tab 42 \tab 67 \tab 121 \tab 272 \tab 137 \tab 129 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 304 \tab 127 \tab 64 \tab 237 \tab 59 \tab 53 \tab 126 \tab 243 \tab 129 \tab 126 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 384 \tab 186 \tab 73 \tab 297 \tab 56 \tab 56 \tab 131 \tab 372 \tab 137 \tab 168 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 344 \tab 125 \tab 76 \tab 271 \tab 63 \tab 92 \tab 124 \tab 217 \tab 161 \tab 126 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 309 \tab 136 \tab 62 \tab 222 \tab 63 \tab 85 \tab 120 \tab 208 \tab 142 \tab 126 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 270 \tab 148 \tab 66 \tab 215 \tab 75 \tab 100 \tab 135 \tab 215 \tab 162 \tab 148 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 306 \tab 133 \tab 72 \tab 222 \tab 73 \tab 85 \tab 122 \tab 195 \tab 145 \tab 117 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 356 \tab 217 \tab 67 \tab 264 \tab 73 \tab 76 \tab 128 \tab 336 \tab 167 \tab 135 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 288 \tab 110 \tab 76 \tab 220 \tab 74 \tab 83 \tab 123 \tab 220 \tab 155 \tab 132 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 312 \tab 116 \tab 82 \tab 237 \tab 69 \tab 75 \tab 151 \tab 230 \tab 147 \tab 118 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 267 \tab 108 \tab 70 \tab 220 \tab 62 \tab 70 \tab 133 \tab 191 \tab 144 \tab 107 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 292 \tab 127 \tab 66 \tab 234 \tab 66 \tab 89 \tab 142 \tab 242 \tab 135 \tab 124 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 332 \tab 137 \tab 84 \tab 244 \tab 66 \tab 57 \tab 172 \tab 223 \tab 136 \tab 155 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 522 \tab 222 \tab 166 \tab 468 \tab 142 \tab 140 \tab 311 \tab 466 \tab 252 \tab 303 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 563 \tab 216 \tab 180 \tab 465 \tab 145 \tab 166 \tab 291 \tab 414 \tab 213 \tab 251 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 523 \tab 203 \tab 174 \tab 452 \tab 143 \tab 145 \tab 260 \tab 433 \tab 225 \tab 274 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 510 \tab 204 \tab 191 \tab 470 \tab 148 \tab 129 \tab 274 \tab 401 \tab 204 \tab 286 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 546 \tab 202 \tab 195 \tab 443 \tab 117 \tab 148 \tab 264 \tab 382 \tab 215 \tab 253 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 548 \tab 151 \tab 194 \tab 398 \tab 160 \tab 105 \tab 312 \tab 459 \tab 190 \tab 295 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 802 \tab 252 \tab 489 \tab 591 \tab 253 \tab 165 \tab 508 \tab 762 \tab 284 \tab 404 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 359 \tab 101 \tab 193 \tab 289 \tab 104 \tab 83 \tab 249 \tab 326 \tab 144 \tab 175 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 380 \tab 130 \tab 189 \tab 294 \tab 97 \tab 76 \tab 244 \tab 321 \tab 134 \tab 179 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 347 \tab 113 \tab 205 \tab 248 \tab 98 \tab 102 \tab 233 \tab 332 \tab 133 \tab 163 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 447 \tab 155 \tab 280 \tab 359 \tab 146 \tab 118 \tab 302 \tab 379 \tab 144 \tab 208 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 552 \tab 187 \tab 249 \tab 371 \tab 147 \tab 159 \tab 339 \tab 435 \tab 185 \tab 243 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 405 \tab 147 \tab 217 \tab 268 \tab 116 \tab 115 \tab 303 \tab 397 \tab 150 \tab 230 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 812 \tab 235 \tab - \tab 543 \tab 242 \tab 176 \tab 783 \tab 536 \tab 335 \tab 356 \tab 14 \tab - \tab - \tab 4032 \cr
   2022 \tab 613 \tab 199 \tab 315 \tab 393 \tab 195 \tab 177 \tab 452 \tab 633 \tab 242 \tab 315 \tab - \tab 1 \tab 9 \tab 3544 \cr
   Total \tab 13905 \tab 5157 \tab 4523 \tab 10781 \tab 3376 \tab 3386 \tab 7591 \tab 11384 \tab 5777 \tab 6486 \tab 14 \tab 1 \tab 9 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} foreign
\item \code{1} new england
\item \code{2} middle atlantic
\item \code{3} east north central
\item \code{4} west north central
\item \code{5} south atlantic
\item \code{6} east south atlantic
\item \code{7} west south central
\item \code{8} mountain
\item \code{9} pacific
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
