% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_17.R
\name{region}
\alias{region}
\title{Region of interview}
\source{
General Social Survey https://gss.norc.org
}
\description{
region
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab east north central \tab east south atlantic \tab middle atlantic \tab mountain \tab new england \tab pacific \tab south atlantic \tab west north central \tab west south central \tab Total \cr
   1972 \tab 305 \tab 84 \tab 298 \tab 56 \tab 90 \tab 224 \tab 218 \tab 141 \tab 197 \tab 1613 \cr
   1973 \tab 333 \tab 75 \tab 268 \tab 59 \tab 75 \tab 194 \tab 284 \tab 95 \tab 121 \tab 1504 \cr
   1974 \tab 334 \tab 75 \tab 257 \tab 55 \tab 71 \tab 195 \tab 282 \tab 95 \tab 120 \tab 1484 \cr
   1975 \tab 329 \tab 84 \tab 262 \tab 61 \tab 64 \tab 174 \tab 295 \tab 100 \tab 121 \tab 1490 \cr
   1976 \tab 321 \tab 85 \tab 270 \tab 64 \tab 70 \tab 190 \tab 282 \tab 103 \tab 114 \tab 1499 \cr
   1977 \tab 371 \tab 76 \tab 239 \tab 61 \tab 62 \tab 204 \tab 313 \tab 85 \tab 119 \tab 1530 \cr
   1978 \tab 327 \tab 85 \tab 263 \tab 66 \tab 72 \tab 186 \tab 305 \tab 116 \tab 112 \tab 1532 \cr
   1980 \tab 306 \tab 111 \tab 239 \tab 79 \tab 53 \tab 191 \tab 267 \tab 102 \tab 120 \tab 1468 \cr
   1982 \tab 402 \tab 149 \tab 331 \tab 58 \tab 63 \tab 220 \tab 379 \tab 110 \tab 148 \tab 1860 \cr
   1983 \tab 330 \tab 101 \tab 254 \tab 88 \tab 85 \tab 200 \tab 252 \tab 154 \tab 135 \tab 1599 \cr
   1984 \tab 282 \tab 118 \tab 217 \tab 86 \tab 78 \tab 181 \tab 249 \tab 134 \tab 128 \tab 1473 \cr
   1985 \tab 270 \tab 123 \tab 201 \tab 102 \tab 94 \tab 190 \tab 264 \tab 140 \tab 150 \tab 1534 \cr
   1986 \tab 270 \tab 116 \tab 217 \tab 88 \tab 86 \tab 188 \tab 252 \tab 134 \tab 119 \tab 1470 \cr
   1987 \tab 362 \tab 150 \tab 263 \tab 103 \tab 78 \tab 189 \tab 389 \tab 144 \tab 141 \tab 1819 \cr
   1988 \tab 268 \tab 105 \tab 209 \tab 105 \tab 80 \tab 176 \tab 265 \tab 139 \tab 134 \tab 1481 \cr
   1989 \tab 256 \tab 114 \tab 221 \tab 92 \tab 82 \tab 212 \tab 293 \tab 145 \tab 122 \tab 1537 \cr
   1990 \tab 243 \tab 105 \tab 200 \tab 85 \tab 73 \tab 179 \tab 238 \tab 134 \tab 115 \tab 1372 \cr
   1991 \tab 262 \tab 114 \tab 207 \tab 79 \tab 92 \tab 215 \tab 301 \tab 118 \tab 129 \tab 1517 \cr
   1993 \tab 299 \tab 129 \tab 230 \tab 86 \tab 60 \tab 254 \tab 278 \tab 118 \tab 152 \tab 1606 \cr
   1994 \tab 484 \tab 203 \tab 416 \tab 202 \tab 150 \tab 414 \tab 575 \tab 231 \tab 317 \tab 2992 \cr
   1996 \tab 489 \tab 209 \tab 412 \tab 200 \tab 157 \tab 420 \tab 542 \tab 210 \tab 265 \tab 2904 \cr
   1998 \tab 492 \tab 189 \tab 428 \tab 188 \tab 141 \tab 358 \tab 532 \tab 204 \tab 300 \tab 2832 \cr
   2000 \tab 467 \tab 195 \tab 429 \tab 182 \tab 139 \tab 391 \tab 510 \tab 203 \tab 301 \tab 2817 \cr
   2002 \tab 478 \tab 199 \tab 435 \tab 165 \tab 141 \tab 366 \tab 486 \tab 223 \tab 272 \tab 2765 \cr
   2004 \tab 494 \tab 161 \tab 363 \tab 203 \tab 86 \tab 379 \tab 611 \tab 203 \tab 312 \tab 2812 \cr
   2006 \tab 762 \tab 255 \tab 548 \tab 364 \tab 163 \tab 652 \tab 1011 \tab 276 \tab 479 \tab 4510 \cr
   2008 \tab 349 \tab 97 \tab 274 \tab 154 \tab 81 \tab 295 \tab 443 \tab 120 \tab 210 \tab 2023 \cr
   2010 \tab 357 \tab 130 \tab 262 \tab 144 \tab 76 \tab 301 \tab 438 \tab 117 \tab 219 \tab 2044 \cr
   2012 \tab 330 \tab 128 \tab 222 \tab 148 \tab 113 \tab 292 \tab 423 \tab 127 \tab 191 \tab 1974 \cr
   2014 \tab 428 \tab 151 \tab 309 \tab 224 \tab 126 \tab 380 \tab 512 \tab 145 \tab 263 \tab 2538 \cr
   2016 \tab 502 \tab 205 \tab 313 \tab 235 \tab 175 \tab 397 \tab 550 \tab 193 \tab 297 \tab 2867 \cr
   2018 \tab 388 \tab 163 \tab 232 \tab 183 \tab 124 \tab 343 \tab 513 \tab 125 \tab 277 \tab 2348 \cr
   2021 \tab 676 \tab 270 \tab 414 \tab 345 \tab 203 \tab 584 \tab 800 \tab 314 \tab 426 \tab 4032 \cr
   2022 \tab 577 \tab 229 \tab 348 \tab 296 \tab 177 \tab 510 \tab 817 \tab 221 \tab 369 \tab 3544 \cr
   Total \tab 13143 \tab 4783 \tab 10051 \tab 4706 \tab 3480 \tab 9844 \tab 14169 \tab 5219 \tab 6995 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} new england
\item \code{2} middle atlantic
\item \code{3} east north central
\item \code{4} west north central
\item \code{5} south atlantic
\item \code{6} east south atlantic
\item \code{7} west south central
\item \code{8} mountain
\item \code{9} pacific
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
