% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_513.R
\name{relactiv}
\alias{relactiv}
\title{relactiv}
\source{
General Social Survey https://gss.norc.org
}
\description{
How often does r take part in relig activities
}
\details{
Question 1394. How often do you take part in the activities and organizations of a church or place of worship other than attending services?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab 2-3 times a month \tab about once a month \tab about once or twice a year \tab every week \tab less than once a year \tab nearly every week \tab never \tab no answer \tab several times a week \tab several times a year \tab don't know \tab once a day \tab several times a day \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 158 \tab 82 \tab 83 \tab 171 \tab 89 \tab 160 \tab 61 \tab 507 \tab 30 \tab 48 \tab 128 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1548 \tab 66 \tab 64 \tab 136 \tab 88 \tab 147 \tab 54 \tab 471 \tab 44 \tab 66 \tab 131 \tab 1 \tab 8 \tab 8 \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 174 \tab 215 \tab 368 \tab 236 \tab 249 \tab 88 \tab 1325 \tab 8 \tab 17 \tab 290 \tab 5 \tab 17 \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 108 \tab 105 \tab 300 \tab 143 \tab 148 \tab 43 \tab 861 \tab 10 \tab 13 \tab 270 \tab 7 \tab 15 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 140 \tab 120 \tab 292 \tab 137 \tab 164 \tab 50 \tab 864 \tab 7 \tab 21 \tab 238 \tab 6 \tab 5 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 129 \tab 118 \tab 251 \tab 167 \tab 139 \tab 64 \tab 887 \tab 12 \tab 9 \tab 183 \tab 3 \tab 12 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 140 \tab 146 \tab 353 \tab 183 \tab 197 \tab 76 \tab 1139 \tab 13 \tab 21 \tab 261 \tab 1 \tab 8 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 162 \tab 183 \tab 384 \tab 213 \tab 187 \tab 69 \tab 1286 \tab 13 \tab 21 \tab 327 \tab 6 \tab 16 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 145 \tab 154 \tab 304 \tab 180 \tab 148 \tab 37 \tab 1087 \tab 10 \tab 13 \tab 241 \tab 20 \tab 9 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 159 \tab 150 \tab 488 \tab 241 \tab 614 \tab 137 \tab 1650 \tab 49 \tab 26 \tab 449 \tab 2 \tab 35 \tab - \tab 32 \tab 4032 \cr
   2022 \tab - \tab 145 \tab 164 \tab 438 \tab 207 \tab 413 \tab 85 \tab 1635 \tab 21 \tab 10 \tab 365 \tab 16 \tab 28 \tab - \tab 17 \tab 3544 \cr
   Total \tab 45385 \tab 1450 \tab 1502 \tab 3485 \tab 1884 \tab 2566 \tab 764 \tab 11712 \tab 217 \tab 265 \tab 2883 \tab 67 \tab 153 \tab 8 \tab 49 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} less than once a year
\item \code{3} about once or twice a year
\item \code{4} several times a year
\item \code{5} about once a month
\item \code{6} 2-3 times a month
\item \code{7} nearly every week
\item \code{8} every week
\item \code{9} several times a week
\item \code{10} once a day
\item \code{11} several times a day
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
