% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_588.R
\name{relate12}
\alias{relate12}
\title{Relation of 12th person (visitor) to head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relate12
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. a. What is \link{PERSON}'s relationship to the head of household?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab child (natural, adopted, step) \tab grand/great-grandchild \tab non-relative \tab other relative \tab parent/parent-in-law \tab son-/daughter-in-law \tab no answer \tab spouse \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1479 \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1490 \cr
   1976 \tab 1495 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1527 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1530 \cr
   1978 \tab 1521 \tab 3 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1532 \cr
   1980 \tab 1457 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 1468 \cr
   1982 \tab 1855 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1860 \cr
   1983 \tab 1593 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1599 \cr
   1984 \tab 1463 \tab 2 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 1 \tab - \tab 1473 \cr
   1985 \tab 1528 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1466 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1470 \cr
   1987 \tab 1812 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1819 \cr
   1988 \tab 1478 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1531 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1369 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1372 \cr
   1991 \tab 1512 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1517 \cr
   1993 \tab 1605 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2987 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 2992 \cr
   1996 \tab 2896 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2904 \cr
   1998 \tab 2817 \tab 7 \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 2813 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2817 \cr
   2002 \tab 2764 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4498 \tab 2 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 4510 \cr
   2008 \tab 2019 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2035 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 2 \tab - \tab 2044 \cr
   2012 \tab 1971 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2535 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2862 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2867 \cr
   2018 \tab 2343 \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 72220 \tab 33 \tab 29 \tab 31 \tab 26 \tab 23 \tab 15 \tab 12 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} child (natural, adopted, step)
\item \code{4} son-/daughter-in-law
\item \code{5} grand/great-grandchild
\item \code{6} parent/parent-in-law
\item \code{7} other relative
\item \code{8} non-relative
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
