% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_588.R
\name{relate13}
\alias{relate13}
\title{Relation of 13th person (visitor) to head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relate13
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. a. What is (PERSON)'s relationship to the head of household?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab grand/great-grandchild \tab other relative \tab child (natural, adopted, step) \tab no answer \tab son-/daughter-in-law \tab spouse \tab non-relative \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1487 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1498 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1529 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1527 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1464 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1858 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1860 \cr
   1983 \tab 1596 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1470 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1473 \cr
   1985 \tab 1533 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1816 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1480 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1536 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2991 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2903 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2826 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 2832 \cr
   2000 \tab 2816 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4508 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2022 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2042 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1973 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2536 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2538 \cr
   2016 \tab 2865 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 72343 \tab 22 \tab 7 \tab 7 \tab 6 \tab 1 \tab 1 \tab 3 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} child (natural, adopted, step)
\item \code{4} son-/daughter-in-law
\item \code{5} grand/great-grandchild
\item \code{6} parent/parent-in-law
\item \code{7} other relative
\item \code{8} non-relative
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
