% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_582.R
\name{relate3}
\alias{relate3}
\title{relate3}
\source{
General Social Survey https://gss.norc.org
}
\description{
Relationship of 3rd person to household head
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. a. What is \link{PERSON}'s relationship to the head of household?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab child (natural, adopted, step) \tab grand/great-grandchild \tab non-relative \tab other relative \tab parent/parent-in-law \tab son-/daughter-in-law \tab spouse \tab no answer \tab head of household \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 656 \tab 768 \tab 7 \tab 22 \tab 17 \tab 11 \tab 6 \tab 3 \tab - \tab - \tab 1490 \cr
   1976 \tab 1134 \tab 333 \tab 7 \tab 8 \tab 9 \tab 5 \tab 1 \tab 2 \tab - \tab - \tab 1499 \cr
   1977 \tab 724 \tab 742 \tab 11 \tab 9 \tab 16 \tab 18 \tab 2 \tab 3 \tab 5 \tab - \tab 1530 \cr
   1978 \tab 749 \tab 714 \tab 9 \tab 17 \tab 18 \tab 11 \tab 4 \tab 3 \tab 7 \tab - \tab 1532 \cr
   1980 \tab 780 \tab 597 \tab 10 \tab 24 \tab 14 \tab 11 \tab 8 \tab - \tab 24 \tab - \tab 1468 \cr
   1982 \tab 991 \tab 766 \tab 27 \tab 20 \tab 16 \tab 18 \tab 5 \tab 5 \tab 12 \tab - \tab 1860 \cr
   1983 \tab 797 \tab 723 \tab 14 \tab 17 \tab 21 \tab 10 \tab 5 \tab 7 \tab 3 \tab 2 \tab 1599 \cr
   1984 \tab 783 \tab 619 \tab 10 \tab 25 \tab 11 \tab 11 \tab 1 \tab 5 \tab 8 \tab - \tab 1473 \cr
   1985 \tab 844 \tab 625 \tab 5 \tab 22 \tab 10 \tab 13 \tab 2 \tab 9 \tab 4 \tab - \tab 1534 \cr
   1986 \tab 768 \tab 612 \tab 12 \tab 29 \tab 15 \tab 14 \tab 6 \tab 8 \tab 6 \tab - \tab 1470 \cr
   1987 \tab 948 \tab 707 \tab 27 \tab 47 \tab 28 \tab 11 \tab 2 \tab 10 \tab 37 \tab 2 \tab 1819 \cr
   1988 \tab 847 \tab 548 \tab 14 \tab 22 \tab 27 \tab 9 \tab 2 \tab 3 \tab 9 \tab - \tab 1481 \cr
   1989 \tab 834 \tab 592 \tab 12 \tab 29 \tab 11 \tab 4 \tab 6 \tab 8 \tab 40 \tab 1 \tab 1537 \cr
   1990 \tab 810 \tab 491 \tab 10 \tab 23 \tab 17 \tab 6 \tab 5 \tab 4 \tab 6 \tab - \tab 1372 \cr
   1991 \tab 849 \tab 577 \tab 10 \tab 34 \tab 22 \tab 10 \tab 1 \tab 5 \tab 9 \tab - \tab 1517 \cr
   1993 \tab 895 \tab 630 \tab 9 \tab 28 \tab 17 \tab 11 \tab 7 \tab 5 \tab 4 \tab - \tab 1606 \cr
   1994 \tab 1751 \tab 1094 \tab 23 \tab 47 \tab 32 \tab 21 \tab 4 \tab 13 \tab 7 \tab - \tab 2992 \cr
   1996 \tab 1729 \tab 970 \tab 21 \tab 63 \tab 24 \tab 20 \tab 10 \tab 13 \tab 54 \tab - \tab 2904 \cr
   1998 \tab 1713 \tab 923 \tab 15 \tab 56 \tab 32 \tab 12 \tab 10 \tab 23 \tab 48 \tab - \tab 2832 \cr
   2000 \tab 1668 \tab 924 \tab 33 \tab 56 \tab 37 \tab 29 \tab 10 \tab - \tab 60 \tab - \tab 2817 \cr
   2002 \tab 1815 \tab 774 \tab 39 \tab 63 \tab 40 \tab 13 \tab 15 \tab - \tab 6 \tab - \tab 2765 \cr
   2004 \tab 1780 \tab 899 \tab 34 \tab 31 \tab 29 \tab 24 \tab 11 \tab - \tab 4 \tab - \tab 2812 \cr
   2006 \tab 2778 \tab 1495 \tab 44 \tab 84 \tab 52 \tab 45 \tab 9 \tab - \tab 3 \tab - \tab 4510 \cr
   2008 \tab 1221 \tab 684 \tab 34 \tab 29 \tab 33 \tab 18 \tab 1 \tab - \tab 3 \tab - \tab 2023 \cr
   2010 \tab 1319 \tab 600 \tab 26 \tab 45 \tab 26 \tab 15 \tab 10 \tab - \tab 3 \tab - \tab 2044 \cr
   2012 \tab 1259 \tab 583 \tab 14 \tab 41 \tab 47 \tab 20 \tab 3 \tab - \tab 7 \tab - \tab 1974 \cr
   2014 \tab 1642 \tab 743 \tab 21 \tab 68 \tab 29 \tab 22 \tab 8 \tab - \tab 5 \tab - \tab 2538 \cr
   2016 \tab 1919 \tab 765 \tab 37 \tab 59 \tab 37 \tab 33 \tab 14 \tab - \tab 3 \tab - \tab 2867 \cr
   2018 \tab 1561 \tab 651 \tab 28 \tab 46 \tab 31 \tab 23 \tab 7 \tab - \tab 1 \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 47741 \tab 21149 \tab 563 \tab 1064 \tab 718 \tab 468 \tab 175 \tab 129 \tab 378 \tab 5 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} child (natural, adopted, step)
\item \code{4} son-/daughter-in-law
\item \code{5} grand/great-grandchild
\item \code{6} parent/parent-in-law
\item \code{7} other relative
\item \code{8} non-relative
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
