% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_583.R
\name{relate4}
\alias{relate4}
\title{Relationship of 4th person to household head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relate4
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. a. What is \link{PERSON}'s relationship to the head of household?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab child (natural, adopted, step) \tab grand/great-grandchild \tab non-relative \tab other relative \tab parent/parent-in-law \tab son-/daughter-in-law \tab spouse \tab no answer \tab head of household \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 952 \tab 496 \tab 13 \tab 12 \tab 10 \tab 5 \tab 2 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1255 \tab 231 \tab 8 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1007 \tab 482 \tab 17 \tab 5 \tab 12 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab 1530 \cr
   1978 \tab 1017 \tab 478 \tab 11 \tab 6 \tab 7 \tab 4 \tab 3 \tab 1 \tab 5 \tab - \tab 1532 \cr
   1980 \tab 1040 \tab 376 \tab 14 \tab 10 \tab 6 \tab 2 \tab 2 \tab - \tab 18 \tab - \tab 1468 \cr
   1982 \tab 1343 \tab 446 \tab 32 \tab 6 \tab 8 \tab 6 \tab 10 \tab 3 \tab 6 \tab - \tab 1860 \cr
   1983 \tab 1104 \tab 451 \tab 11 \tab 5 \tab 11 \tab 2 \tab 9 \tab 1 \tab 2 \tab 3 \tab 1599 \cr
   1984 \tab 1062 \tab 374 \tab 16 \tab 6 \tab 4 \tab 2 \tab 2 \tab 2 \tab 5 \tab - \tab 1473 \cr
   1985 \tab 1112 \tab 384 \tab 8 \tab 13 \tab 7 \tab 5 \tab - \tab 2 \tab 2 \tab 1 \tab 1534 \cr
   1986 \tab 1032 \tab 388 \tab 18 \tab 13 \tab 6 \tab 2 \tab 4 \tab 2 \tab 5 \tab - \tab 1470 \cr
   1987 \tab 1282 \tab 418 \tab 35 \tab 28 \tab 16 \tab 4 \tab 5 \tab 3 \tab 28 \tab - \tab 1819 \cr
   1988 \tab 1097 \tab 346 \tab 13 \tab 7 \tab 9 \tab 3 \tab 1 \tab 2 \tab 3 \tab - \tab 1481 \cr
   1989 \tab 1098 \tab 370 \tab 16 \tab 14 \tab 4 \tab 3 \tab 4 \tab 7 \tab 21 \tab - \tab 1537 \cr
   1990 \tab 1032 \tab 295 \tab 13 \tab 12 \tab 6 \tab 2 \tab 6 \tab 1 \tab 5 \tab - \tab 1372 \cr
   1991 \tab 1125 \tab 338 \tab 13 \tab 17 \tab 17 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 1517 \cr
   1993 \tab 1190 \tab 371 \tab 15 \tab 15 \tab 6 \tab 2 \tab 2 \tab 3 \tab 2 \tab - \tab 1606 \cr
   1994 \tab 2274 \tab 624 \tab 34 \tab 26 \tab 15 \tab 6 \tab 5 \tab 5 \tab 3 \tab - \tab 2992 \cr
   1996 \tab 2184 \tab 591 \tab 26 \tab 27 \tab 23 \tab 6 \tab 6 \tab 10 \tab 31 \tab - \tab 2904 \cr
   1998 \tab 2195 \tab 534 \tab 23 \tab 19 \tab 18 \tab 6 \tab 8 \tab 5 \tab 24 \tab - \tab 2832 \cr
   2000 \tab 2127 \tab 569 \tab 35 \tab 21 \tab 16 \tab 7 \tab 6 \tab - \tab 36 \tab - \tab 2817 \cr
   2002 \tab 2266 \tab 417 \tab 30 \tab 23 \tab 16 \tab 1 \tab 7 \tab - \tab 5 \tab - \tab 2765 \cr
   2004 \tab 2214 \tab 522 \tab 31 \tab 17 \tab 16 \tab 5 \tab 5 \tab - \tab 2 \tab - \tab 2812 \cr
   2006 \tab 3502 \tab 897 \tab 33 \tab 37 \tab 27 \tab 10 \tab 3 \tab - \tab 1 \tab - \tab 4510 \cr
   2008 \tab 1547 \tab 405 \tab 34 \tab 18 \tab 11 \tab 2 \tab 4 \tab - \tab 2 \tab - \tab 2023 \cr
   2010 \tab 1627 \tab 329 \tab 28 \tab 27 \tab 14 \tab 6 \tab 10 \tab - \tab 3 \tab - \tab 2044 \cr
   2012 \tab 1526 \tab 361 \tab 28 \tab 21 \tab 26 \tab 6 \tab 4 \tab - \tab 2 \tab - \tab 1974 \cr
   2014 \tab 2040 \tab 420 \tab 24 \tab 23 \tab 11 \tab 7 \tab 11 \tab - \tab 2 \tab - \tab 2538 \cr
   2016 \tab 2353 \tab 426 \tab 34 \tab 16 \tab 17 \tab 8 \tab 10 \tab - \tab 3 \tab - \tab 2867 \cr
   2018 \tab 1899 \tab 354 \tab 29 \tab 28 \tab 17 \tab 5 \tab 16 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 57679 \tab 12693 \tab 642 \tab 473 \tab 359 \tab 124 \tab 150 \tab 48 \tab 218 \tab 4 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} child (natural, adopted, step)
\item \code{4} son-/daughter-in-law
\item \code{5} grand/great-grandchild
\item \code{6} parent/parent-in-law
\item \code{7} other relative
\item \code{8} non-relative
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
