% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_584.R
\name{relate5}
\alias{relate5}
\title{Relationship of 5th person to household head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relate5
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. a. What is \link{PERSON}'s relationship to the head of household?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab child (natural, adopted, step) \tab grand/great-grandchild \tab non-relative \tab other relative \tab parent/parent-in-law \tab spouse \tab no answer \tab son-/daughter-in-law \tab head of household \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1202 \tab 263 \tab 10 \tab 4 \tab 5 \tab 2 \tab 4 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1369 \tab 117 \tab 9 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1264 \tab 247 \tab 6 \tab 2 \tab 3 \tab 4 \tab - \tab 1 \tab 3 \tab - \tab 1530 \cr
   1978 \tab 1293 \tab 223 \tab 5 \tab 1 \tab 3 \tab 4 \tab - \tab 3 \tab - \tab - \tab 1532 \cr
   1980 \tab 1283 \tab 163 \tab 4 \tab 3 \tab 1 \tab 2 \tab - \tab 11 \tab 1 \tab - \tab 1468 \cr
   1982 \tab 1610 \tab 203 \tab 30 \tab 3 \tab 4 \tab 4 \tab 2 \tab 3 \tab 1 \tab - \tab 1860 \cr
   1983 \tab 1376 \tab 190 \tab 17 \tab 2 \tab 12 \tab - \tab - \tab - \tab - \tab 2 \tab 1599 \cr
   1984 \tab 1280 \tab 168 \tab 11 \tab 2 \tab 3 \tab 3 \tab - \tab 3 \tab 2 \tab 1 \tab 1473 \cr
   1985 \tab 1347 \tab 173 \tab 5 \tab 4 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1534 \cr
   1986 \tab 1265 \tab 170 \tab 17 \tab 2 \tab 8 \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab 1470 \cr
   1987 \tab 1577 \tab 168 \tab 30 \tab 13 \tab 10 \tab - \tab 2 \tab 19 \tab - \tab - \tab 1819 \cr
   1988 \tab 1328 \tab 133 \tab 6 \tab 5 \tab 6 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1356 \tab 143 \tab 13 \tab 9 \tab 1 \tab 1 \tab - \tab 12 \tab 2 \tab - \tab 1537 \cr
   1990 \tab 1237 \tab 116 \tab 9 \tab 1 \tab 4 \tab 1 \tab - \tab 4 \tab - \tab - \tab 1372 \cr
   1991 \tab 1367 \tab 120 \tab 15 \tab 5 \tab 8 \tab - \tab - \tab 2 \tab - \tab - \tab 1517 \cr
   1993 \tab 1451 \tab 128 \tab 12 \tab 6 \tab 7 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1606 \cr
   1994 \tab 2729 \tab 209 \tab 26 \tab 13 \tab 9 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 2992 \cr
   1996 \tab 2637 \tab 204 \tab 19 \tab 10 \tab 18 \tab 4 \tab 2 \tab 10 \tab - \tab - \tab 2904 \cr
   1998 \tab 2588 \tab 196 \tab 9 \tab 9 \tab 9 \tab 4 \tab 1 \tab 14 \tab 2 \tab - \tab 2832 \cr
   2000 \tab 2559 \tab 197 \tab 20 \tab 10 \tab 13 \tab 3 \tab - \tab 15 \tab - \tab - \tab 2817 \cr
   2002 \tab 2556 \tab 172 \tab 18 \tab 8 \tab 9 \tab - \tab - \tab 2 \tab - \tab - \tab 2765 \cr
   2004 \tab 2567 \tab 198 \tab 24 \tab 8 \tab 9 \tab 4 \tab - \tab - \tab 2 \tab - \tab 2812 \cr
   2006 \tab 4086 \tab 364 \tab 26 \tab 15 \tab 13 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 4510 \cr
   2008 \tab 1826 \tab 162 \tab 22 \tab 7 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab 1861 \tab 133 \tab 25 \tab 7 \tab 12 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 2044 \cr
   2012 \tab 1775 \tab 154 \tab 26 \tab 9 \tab 7 \tab 3 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2325 \tab 176 \tab 18 \tab 7 \tab 9 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab 2632 \tab 176 \tab 30 \tab 11 \tab 14 \tab 3 \tab - \tab - \tab 1 \tab - \tab 2867 \cr
   2018 \tab 2192 \tab 110 \tab 25 \tab 9 \tab 5 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 66115 \tab 5176 \tab 487 \tab 185 \tab 211 \tab 58 \tab 16 \tab 113 \tab 26 \tab 3 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} child (natural, adopted, step)
\item \code{4} son-/daughter-in-law
\item \code{5} grand/great-grandchild
\item \code{6} parent/parent-in-law
\item \code{7} other relative
\item \code{8} non-relative
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
