% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_585.R
\name{relate7}
\alias{relate7}
\title{Relationship of 7th person to household head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relate7
}
\details{
Question 1621. Please tell me the names of the people who usually live in this household. Let's start with the head of the household. a. What is (PERSON)'s relationship to the head of household?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab child (natural, adopted, step) \tab grand/great-grandchild \tab non-relative \tab parent/parent-in-law \tab no answer \tab other relative \tab spouse \tab son-/daughter-in-law \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1437 \tab 46 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1479 \tab 19 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1484 \tab 40 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1530 \cr
   1978 \tab 1492 \tab 33 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 1532 \cr
   1980 \tab 1431 \tab 29 \tab 5 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1817 \tab 31 \tab 9 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1559 \tab 38 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1454 \tab 11 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1473 \cr
   1985 \tab 1505 \tab 22 \tab 5 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1441 \tab 21 \tab 6 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1470 \cr
   1987 \tab 1787 \tab 18 \tab 7 \tab 5 \tab - \tab - \tab 2 \tab - \tab - \tab 1819 \cr
   1988 \tab 1465 \tab 12 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1481 \cr
   1989 \tab 1498 \tab 23 \tab 8 \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab 1537 \cr
   1990 \tab 1354 \tab 14 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1372 \cr
   1991 \tab 1497 \tab 16 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1517 \cr
   1993 \tab 1597 \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1606 \cr
   1994 \tab 2958 \tab 19 \tab 10 \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 2992 \cr
   1996 \tab 2862 \tab 21 \tab 7 \tab 4 \tab 2 \tab 4 \tab 4 \tab - \tab - \tab 2904 \cr
   1998 \tab 2801 \tab 14 \tab 4 \tab 1 \tab - \tab 2 \tab 9 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 2786 \tab 22 \tab 2 \tab - \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 2817 \cr
   2002 \tab 2741 \tab 14 \tab 7 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab 2786 \tab 16 \tab 3 \tab 3 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2812 \cr
   2006 \tab 4457 \tab 40 \tab 7 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab - \tab 4510 \cr
   2008 \tab 2002 \tab 14 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2003 \tab 22 \tab 10 \tab 1 \tab - \tab 2 \tab 6 \tab - \tab - \tab 2044 \cr
   2012 \tab 1951 \tab 13 \tab 6 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1974 \cr
   2014 \tab 2515 \tab 16 \tab 4 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 2538 \cr
   2016 \tab 2830 \tab 28 \tab 7 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2867 \cr
   2018 \tab 2327 \tab 10 \tab 6 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 71493 \tab 629 \tab 140 \tab 33 \tab 14 \tab 21 \tab 52 \tab 5 \tab 3 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} child (natural, adopted, step)
\item \code{4} son-/daughter-in-law
\item \code{5} grand/great-grandchild
\item \code{6} parent/parent-in-law
\item \code{7} other relative
\item \code{8} non-relative
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
