% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_553.R
\name{relatsex}
\alias{relatsex}
\title{relatsex}
\source{
General Social Survey https://gss.norc.org
}
\description{
Relation to last sex partner
}
\details{
Question 1557. The last time you had sex, was it with someone you were in an on-going relationship with, or was it with someone else? Remember that by "sex" we mean only vaginal, oral, or anal sex.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no answer \tab no, the last time i had sex, it was not with someone i was in an on-going relationship with \tab yes, the last time i had sex, it was with someone i was in an on-going relationship with \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 258 \tab 190 \tab 196 \tab 2260 \tab - \tab - \tab 2904 \cr
   1998 \tab 383 \tab 162 \tab 236 \tab 2048 \tab 3 \tab - \tab 2832 \cr
   2000 \tab 417 \tab 204 \tab 236 \tab 1959 \tab 1 \tab - \tab 2817 \cr
   2002 \tab 476 \tab 139 \tab 158 \tab 1980 \tab 12 \tab - \tab 2765 \cr
   2004 \tab 570 \tab 96 \tab 176 \tab 1956 \tab 14 \tab - \tab 2812 \cr
   2006 \tab 2096 \tab 97 \tab 229 \tab 2074 \tab 14 \tab - \tab 4510 \cr
   2008 \tab 240 \tab 47 \tab 167 \tab 1560 \tab 9 \tab - \tab 2023 \cr
   2010 \tab 202 \tab 65 \tab 157 \tab 1611 \tab 9 \tab - \tab 2044 \cr
   2012 \tab 238 \tab 79 \tab 173 \tab 1476 \tab 8 \tab - \tab 1974 \cr
   2014 \tab 188 \tab 92 \tab 249 \tab 1999 \tab 10 \tab - \tab 2538 \cr
   2016 \tab 1096 \tab 68 \tab 219 \tab 1476 \tab 8 \tab - \tab 2867 \cr
   2018 \tab 942 \tab 61 \tab 152 \tab 1187 \tab 6 \tab - \tab 2348 \cr
   2021 \tab 1706 \tab 6 \tab 193 \tab 1993 \tab 89 \tab 45 \tab 4032 \cr
   2022 \tab 1900 \tab 14 \tab 174 \tab 1280 \tab 103 \tab 73 \tab 3544 \cr
   Total \tab 43092 \tab 1320 \tab 2715 \tab 24859 \tab 286 \tab 118 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes, the last time i had sex, it was with someone i was in an on-going relationship with
\item \code{2} no, the last time i had sex, it was not with someone i was in an on-going relationship with
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
