% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_592.R
\name{relhh2}
\alias{relhh2}
\title{Relationship of person 2 to head of household}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhh2
}
\details{
Question 1632. What is (PERSON)'s relationship to head of household?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllll}{
   year \tab iap \tab adopted child \tab brother, sister (full, half, or step sibling) \tab child, not specified \tab cousin \tab friend \tab grandchild, great-grandchild \tab grandparent, great-grandparent \tab natural/biological child \tab niece or nephew \tab no answer \tab other relative (great-aunt, grandniece, etc.) \tab parent \tab parent-in-law \tab partner, fiance(e), boyfriend, girlfriend, etc. \tab roommate, housemate \tab sibling-in-law \tab son-in-law, daughter-in-law \tab spouse \tab stepchild \tab uncle or aunt \tab child of non-relative \tab don't know \tab other non-relative (employee, boarder) \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 741 \tab 2 \tab 29 \tab 65 \tab 3 \tab 13 \tab 13 \tab 1 \tab 284 \tab 9 \tab 10 \tab 1 \tab 26 \tab 1 \tab 192 \tab 92 \tab 1 \tab 4 \tab 1327 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 883 \tab 3 \tab 36 \tab 93 \tab 4 \tab 24 \tab 7 \tab - \tab 228 \tab 1 \tab 6 \tab - \tab 11 \tab - \tab 161 \tab 51 \tab 2 \tab 13 \tab 1215 \tab 4 \tab 1 \tab 2 \tab 1 \tab 19 \tab - \tab 2765 \cr
   2004 \tab 707 \tab 4 \tab 27 \tab 87 \tab 3 \tab 8 \tab 13 \tab 1 \tab 186 \tab 5 \tab 1 \tab - \tab 22 \tab - \tab 163 \tab 44 \tab 2 \tab 5 \tab 1527 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 2812 \cr
   2006 \tab 1226 \tab 3 \tab 40 \tab 171 \tab 6 \tab 28 \tab 18 \tab 2 \tab 287 \tab 10 \tab 2 \tab 1 \tab 30 \tab - \tab 301 \tab 80 \tab 1 \tab 4 \tab 2289 \tab 3 \tab 1 \tab - \tab 2 \tab 5 \tab - \tab 4510 \cr
   2008 \tab 527 \tab 2 \tab 25 \tab 115 \tab 4 \tab 10 \tab 15 \tab - \tab 110 \tab 4 \tab 1 \tab - \tab 11 \tab 1 \tab 151 \tab 31 \tab - \tab 5 \tab 1002 \tab 1 \tab 2 \tab - \tab 3 \tab 3 \tab - \tab 2023 \cr
   2010 \tab 595 \tab 1 \tab 17 \tab 148 \tab - \tab 12 \tab 10 \tab - \tab 75 \tab 7 \tab 9 \tab - \tab 18 \tab 1 \tab 192 \tab 36 \tab - \tab 4 \tab 909 \tab 1 \tab 1 \tab - \tab 5 \tab 3 \tab - \tab 2044 \cr
   2012 \tab 533 \tab 1 \tab 21 \tab 115 \tab 3 \tab 10 \tab 15 \tab - \tab 70 \tab 7 \tab 9 \tab - \tab 23 \tab - \tab 173 \tab 50 \tab 1 \tab 1 \tab 937 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 678 \tab 2 \tab 26 \tab 153 \tab 3 \tab 7 \tab 13 \tab 2 \tab 135 \tab 7 \tab 6 \tab - \tab 17 \tab - \tab 218 \tab 53 \tab - \tab 8 \tab 1198 \tab - \tab 3 \tab 1 \tab 5 \tab 3 \tab - \tab 2538 \cr
   2016 \tab 849 \tab 2 \tab 39 \tab 236 \tab 3 \tab 14 \tab 16 \tab 2 \tab 73 \tab 5 \tab 3 \tab - \tab 22 \tab - \tab 269 \tab 67 \tab - \tab 3 \tab 1256 \tab 4 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2867 \cr
   2018 \tab 677 \tab 1 \tab 19 \tab 148 \tab 3 \tab 4 \tab 11 \tab - \tab 99 \tab 7 \tab 3 \tab 1 \tab 22 \tab 2 \tab 217 \tab 64 \tab 6 \tab 1 \tab 1057 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 49076 \tab 21 \tab 279 \tab 1331 \tab 32 \tab 130 \tab 131 \tab 8 \tab 1547 \tab 62 \tab 50 \tab 3 \tab 202 \tab 5 \tab 2037 \tab 568 \tab 13 \tab 48 \tab 12717 \tab 17 \tab 12 \tab 5 \tab 25 \tab 39 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} partner, fiance(e), boyfriend, girlfriend, etc.
\item \code{4} child, not specified
\item \code{5} grandchild, great-grandchild
\item \code{6} son-in-law, daughter-in-law
\item \code{7} parent
\item \code{8} parent-in-law
\item \code{9} grandparent, great-grandparent
\item \code{10} uncle or aunt
\item \code{11} niece or nephew
\item \code{12} cousin
\item \code{13} brother, sister (full, half, or step sibling)
\item \code{14} sibling-in-law
\item \code{15} other relative (great-aunt, grandniece, etc.)
\item \code{16} roommate, housemate
\item \code{17} friend
\item \code{18} child of non-relative
\item \code{19} other non-relative (employee, boarder)
\item \code{41} natural/biological child
\item \code{42} adopted child
\item \code{43} stepchild
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
