% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_590.R
\name{relhhd1}
\alias{relhhd1}
\title{Relation of 1st person to household head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhhd1
}
\details{
Question 1622. What is (PERSON)'s relationship to (HEAD OF HOUSEHOLD)?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab head of household \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 1606 \tab - \tab 1606 \cr
   1994 \tab - \tab 2992 \tab - \tab 2992 \cr
   1996 \tab - \tab 2904 \tab - \tab 2904 \cr
   1998 \tab - \tab 2832 \tab - \tab 2832 \cr
   2000 \tab - \tab 2817 \tab - \tab 2817 \cr
   2002 \tab - \tab 2765 \tab - \tab 2765 \cr
   2004 \tab - \tab 2812 \tab - \tab 2812 \cr
   2006 \tab - \tab 4510 \tab - \tab 4510 \cr
   2008 \tab - \tab 2023 \tab - \tab 2023 \cr
   2010 \tab - \tab 2044 \tab - \tab 2044 \cr
   2012 \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab 3544 \cr
   Total \tab 31326 \tab 37032 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} partner, fiance(e), boyfriend, girlfriend, etc.
\item \code{4} child, natural or adopted, stepchild
\item \code{5} grandchild, great-grandchild
\item \code{6} son-in-law, daughter-in-law
\item \code{7} parent
\item \code{8} parent-in-law
\item \code{9} grandparent, great-grandparent
\item \code{10} uncle or aunt
\item \code{11} niece or nephew
\item \code{12} cousin
\item \code{13} brother, sister (full, half, or step sibling)
\item \code{14} sibling-in-law
\item \code{15} other relative (great-aunt, grandniece, etc.)
\item \code{16} roommate, housemate
\item \code{17} friend
\item \code{18} child of non-relative
\item \code{19} other non-relative (employee, boarder)
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
