% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_590.R
\name{relhhd4}
\alias{relhhd4}
\title{relhhd4}
\source{
General Social Survey https://gss.norc.org
}
\description{
Relation of 4th person to household head
}
\details{
Question 1622. What is (PERSON)'s relationship to (HEAD OF HOUSEHOLD)?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab brother, sister (full, half, or step sibling) \tab child of non-relative \tab child, natural or adopted, stepchild \tab friend \tab grandchild, great-grandchild \tab niece or nephew \tab no answer \tab other non-relative (employee, boarder) \tab parent \tab parent-in-law \tab partner, fiance(e), boyfriend, girlfriend, etc. \tab roommate, housemate \tab sibling-in-law \tab son-in-law, daughter-in-law \tab spouse \tab cousin \tab other relative (great-aunt, grandniece, etc.) \tab grandparent, great-grandparent \tab uncle or aunt \tab don't know \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1190 \tab 1 \tab 3 \tab 371 \tab 2 \tab 15 \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2274 \tab 2 \tab 11 \tab 624 \tab 3 \tab 34 \tab 5 \tab 3 \tab 1 \tab 3 \tab 3 \tab 2 \tab 9 \tab 4 \tab 5 \tab 5 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2184 \tab 6 \tab 10 \tab 591 \tab 2 \tab 26 \tab 9 \tab 31 \tab 3 \tab 3 \tab 3 \tab 3 \tab 9 \tab 3 \tab 6 \tab 10 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2904 \cr
   1998 \tab 2196 \tab 3 \tab 7 \tab 534 \tab - \tab 23 \tab 10 \tab 24 \tab 2 \tab 4 \tab 2 \tab 4 \tab 6 \tab 2 \tab 7 \tab 5 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2127 \tab 5 \tab 5 \tab 569 \tab 3 \tab 35 \tab 1 \tab 35 \tab 2 \tab 5 \tab 2 \tab - \tab 11 \tab 4 \tab 6 \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 2817 \cr
   2002 \tab 2266 \tab 7 \tab 4 \tab 417 \tab 6 \tab 30 \tab 8 \tab 4 \tab 5 \tab - \tab 1 \tab - \tab 8 \tab - \tab 7 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2765 \cr
   2004 \tab 2214 \tab 6 \tab 2 \tab 522 \tab 2 \tab 31 \tab 4 \tab - \tab 5 \tab 3 \tab 2 \tab - \tab 8 \tab 1 \tab 5 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 2812 \cr
   2006 \tab 3502 \tab 7 \tab 6 \tab 897 \tab 7 \tab 33 \tab 8 \tab - \tab 8 \tab 7 \tab 3 \tab - \tab 16 \tab 6 \tab 3 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 1547 \tab 5 \tab 4 \tab 405 \tab 3 \tab 34 \tab 4 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 8 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2023 \cr
   2010 \tab 1627 \tab 7 \tab 2 \tab 329 \tab 3 \tab 28 \tab 4 \tab 3 \tab 4 \tab 3 \tab 3 \tab 3 \tab 15 \tab 1 \tab 10 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab 1526 \tab 6 \tab 4 \tab 361 \tab 2 \tab 28 \tab 8 \tab 1 \tab 2 \tab 5 \tab 1 \tab 2 \tab 11 \tab 6 \tab 4 \tab - \tab - \tab - \tab - \tab 6 \tab 1 \tab - \tab 1974 \cr
   2014 \tab 2040 \tab 4 \tab 4 \tab 420 \tab 4 \tab 24 \tab 3 \tab 1 \tab 2 \tab 6 \tab 1 \tab 2 \tab 11 \tab - \tab 11 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 2353 \tab 5 \tab 2 \tab 426 \tab 1 \tab 34 \tab 5 \tab - \tab 3 \tab 6 \tab 2 \tab - \tab 10 \tab 1 \tab 10 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 1899 \tab 3 \tab 3 \tab 354 \tab 2 \tab 29 \tab 5 \tab - \tab 3 \tab 1 \tab 4 \tab - \tab 20 \tab 3 \tab 16 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 60271 \tab 67 \tab 67 \tab 6820 \tab 40 \tab 404 \tab 78 \tab 105 \tab 45 \tab 47 \tab 30 \tab 17 \tab 149 \tab 33 \tab 96 \tab 23 \tab 12 \tab 20 \tab 7 \tab 17 \tab 10 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} partner, fiance(e), boyfriend, girlfriend, etc.
\item \code{4} child, natural or adopted, stepchild
\item \code{5} grandchild, great-grandchild
\item \code{6} son-in-law, daughter-in-law
\item \code{7} parent
\item \code{8} parent-in-law
\item \code{9} grandparent, great-grandparent
\item \code{10} uncle or aunt
\item \code{11} niece or nephew
\item \code{12} cousin
\item \code{13} brother, sister (full, half, or step sibling)
\item \code{14} sibling-in-law
\item \code{15} other relative (great-aunt, grandniece, etc.)
\item \code{16} roommate, housemate
\item \code{17} friend
\item \code{18} child of non-relative
\item \code{19} other non-relative (employee, boarder)
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
