% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_34.R
\name{relig}
\alias{relig}
\title{relig}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's religious preference
}
\details{
Question 104. What is your religous preference? Is it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllll}{
   year \tab catholic \tab jewish \tab no answer \tab none \tab other \tab protestant \tab don't know \tab buddhism \tab christian \tab hinduism \tab inter-nondenominational \tab muslim/islam \tab native american \tab orthodox-christian \tab other eastern religions \tab skipped on web \tab Total \cr
   1972 \tab 413 \tab 54 \tab 5 \tab 83 \tab 27 \tab 1031 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 388 \tab 42 \tab 4 \tab 96 \tab 34 \tab 940 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 376 \tab 44 \tab 1 \tab 101 \tab 8 \tab 954 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 363 \tab 23 \tab 2 \tab 113 \tab 14 \tab 975 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 390 \tab 27 \tab 2 \tab 114 \tab 15 \tab 951 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 373 \tab 35 \tab 7 \tab 93 \tab 18 \tab 1004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 384 \tab 29 \tab 4 \tab 119 \tab 17 \tab 979 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 362 \tab 32 \tab 3 \tab 105 \tab 30 \tab 936 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 399 \tab 38 \tab 12 \tab 132 \tab 21 \tab 1258 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 439 \tab 43 \tab 4 \tab 117 \tab 26 \tab 970 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 375 \tab 27 \tab 12 \tab 107 \tab 20 \tab 932 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 408 \tab 32 \tab 5 \tab 109 \tab 25 \tab 955 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 379 \tab 38 \tab 3 \tab 98 \tab 31 \tab 921 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 376 \tab 21 \tab 7 \tab 121 \tab 40 \tab 1254 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 384 \tab 30 \tab 1 \tab 118 \tab 42 \tab 906 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 386 \tab 23 \tab 4 \tab 120 \tab 34 \tab 970 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 327 \tab 27 \tab 4 \tab 109 \tab 42 \tab 862 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 386 \tab 32 \tab 2 \tab 102 \tab 29 \tab 966 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 351 \tab 33 \tab 8 \tab 146 \tab 42 \tab 1025 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 759 \tab 59 \tab 11 \tab 274 \tab 115 \tab 1774 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 685 \tab 68 \tab 4 \tab 339 \tab 143 \tab 1664 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 705 \tab 50 \tab 30 \tab 396 \tab 28 \tab 1524 \tab 5 \tab 9 \tab 33 \tab 5 \tab 19 \tab 13 \tab 3 \tab 10 \tab 2 \tab - \tab 2832 \cr
   2000 \tab 679 \tab 63 \tab 3 \tab 398 \tab 42 \tab 1521 \tab 1 \tab 17 \tab 39 \tab 8 \tab 17 \tab 12 \tab 4 \tab 12 \tab 1 \tab - \tab 2817 \cr
   2002 \tab 673 \tab 48 \tab 15 \tab 379 \tab 22 \tab 1460 \tab 4 \tab 21 \tab 65 \tab 6 \tab 30 \tab 13 \tab 4 \tab 23 \tab 2 \tab - \tab 2765 \cr
   2004 \tab 656 \tab 55 \tab 12 \tab 403 \tab 37 \tab 1483 \tab - \tab 13 \tab 71 \tab 13 \tab 26 \tab 16 \tab 1 \tab 13 \tab 13 \tab - \tab 2812 \cr
   2006 \tab 1114 \tab 78 \tab 26 \tab 739 \tab 39 \tab 2328 \tab - \tab 31 \tab 102 \tab 11 \tab 6 \tab 17 \tab 3 \tab 14 \tab 2 \tab - \tab 4510 \cr
   2008 \tab 470 \tab 39 \tab 8 \tab 332 \tab 12 \tab 1040 \tab 1 \tab 15 \tab 68 \tab 7 \tab 4 \tab 13 \tab 1 \tab 10 \tab 3 \tab - \tab 2023 \cr
   2010 \tab 482 \tab 37 \tab 12 \tab 363 \tab 20 \tab 973 \tab 1 \tab 18 \tab 90 \tab 4 \tab 20 \tab 11 \tab 2 \tab 8 \tab 3 \tab - \tab 2044 \cr
   2012 \tab 444 \tab 28 \tab 2 \tab 387 \tab 25 \tab 916 \tab 5 \tab 6 \tab 120 \tab 9 \tab 2 \tab 13 \tab 6 \tab 6 \tab 5 \tab - \tab 1974 \cr
   2014 \tab 606 \tab 40 \tab 14 \tab 522 \tab 21 \tab 1126 \tab 6 \tab 27 \tab 132 \tab 16 \tab 4 \tab 10 \tab 2 \tab 9 \tab 3 \tab - \tab 2538 \cr
   2016 \tab 649 \tab 51 \tab 18 \tab 620 \tab 29 \tab 1380 \tab 3 \tab 22 \tab 41 \tab 13 \tab 7 \tab 19 \tab 4 \tab 7 \tab 4 \tab - \tab 2867 \cr
   2018 \tab 493 \tab 39 \tab 16 \tab 542 \tab 32 \tab 1139 \tab 5 \tab 19 \tab 30 \tab 8 \tab 1 \tab 16 \tab 1 \tab 6 \tab 1 \tab - \tab 2348 \cr
   2021 \tab 824 \tab 75 \tab 44 \tab 1121 \tab 55 \tab 1590 \tab 5 \tab 47 \tab 124 \tab 30 \tab 18 \tab 25 \tab 3 \tab 37 \tab 2 \tab 32 \tab 4032 \cr
   2022 \tab 744 \tab 61 \tab 32 \tab 977 \tab 89 \tab 1418 \tab 20 \tab 27 \tab 85 \tab 28 \tab 5 \tab 23 \tab 2 \tab 21 \tab 3 \tab 9 \tab 3544 \cr
   Total \tab 17242 \tab 1421 \tab 337 \tab 9895 \tab 1224 \tab 40125 \tab 59 \tab 272 \tab 1000 \tab 158 \tab 159 \tab 201 \tab 36 \tab 176 \tab 44 \tab 41 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{6} buddhism
\item \code{7} hinduism
\item \code{8} other eastern religions
\item \code{9} muslim/islam
\item \code{10} orthodox-christian
\item \code{11} christian
\item \code{12} native american
\item \code{13} inter-nondenominational
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
