% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_39.R
\name{relig16}
\alias{relig16}
\title{Religion in which raised}
\source{
General Social Survey https://gss.norc.org
}
\description{
relig16
}
\details{
Question 115. In what religion were you raised?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab catholic \tab jewish \tab no answer \tab none \tab other \tab protestant \tab don't know \tab buddhism \tab christian \tab hinduism \tab inter-nondenominational \tab muslim/islam \tab native american \tab orthodox-christian \tab other eastern religions \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 419 \tab 43 \tab 4 \tab 36 \tab 21 \tab 981 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 402 \tab 49 \tab 2 \tab 39 \tab 8 \tab 984 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 378 \tab 25 \tab - \tab 45 \tab 10 \tab 1032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 428 \tab 33 \tab 2 \tab 47 \tab 18 \tab 971 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 390 \tab 37 \tab 6 \tab 39 \tab 15 \tab 1042 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 418 \tab 31 \tab 7 \tab 48 \tab 11 \tab 1017 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 389 \tab 32 \tab 2 \tab 45 \tab 17 \tab 983 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 441 \tab 44 \tab 9 \tab 48 \tab 23 \tab 1295 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 481 \tab 50 \tab 9 \tab 52 \tab 11 \tab 996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 399 \tab 30 \tab 21 \tab 43 \tab 25 \tab 955 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 444 \tab 32 \tab 5 \tab 63 \tab 12 \tab 978 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 411 \tab 33 \tab 2 \tab 45 \tab 26 \tab 953 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 407 \tab 22 \tab 29 \tab 48 \tab 23 \tab 1290 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 414 \tab 31 \tab - \tab 53 \tab 31 \tab 952 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 444 \tab 28 \tab 4 \tab 61 \tab 34 \tab 966 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 361 \tab 27 \tab 8 \tab 72 \tab 27 \tab 877 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 434 \tab 34 \tab 3 \tab 58 \tab 22 \tab 966 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 422 \tab 35 \tab 10 \tab 71 \tab 29 \tab 1039 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 882 \tab 55 \tab 7 \tab 127 \tab 74 \tab 1846 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 837 \tab 68 \tab 5 \tab 169 \tab 91 \tab 1734 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 857 \tab 53 \tab 27 \tab 175 \tab 13 \tab 1637 \tab 3 \tab 6 \tab 21 \tab 5 \tab 2 \tab 15 \tab 5 \tab 10 \tab 3 \tab - \tab 2832 \cr
   2000 \tab - \tab 844 \tab 64 \tab 14 \tab 192 \tab 18 \tab 1610 \tab - \tab 15 \tab 21 \tab 7 \tab 4 \tab 7 \tab 6 \tab 13 \tab 2 \tab - \tab 2817 \cr
   2002 \tab - \tab 821 \tab 56 \tab 15 \tab 195 \tab 8 \tab 1553 \tab 6 \tab 13 \tab 38 \tab 5 \tab 13 \tab 14 \tab 1 \tab 26 \tab 1 \tab - \tab 2765 \cr
   2004 \tab - \tab 829 \tab 60 \tab 6 \tab 226 \tab 8 \tab 1568 \tab 5 \tab 17 \tab 40 \tab 9 \tab 7 \tab 19 \tab 1 \tab 15 \tab 2 \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 960 \tab 56 \tab 1 \tab 249 \tab 5 \tab 1620 \tab - \tab 21 \tab 26 \tab 16 \tab - \tab 23 \tab 2 \tab 9 \tab 4 \tab - \tab 4510 \cr
   2008 \tab - \tab 635 \tab 39 \tab 10 \tab 168 \tab 4 \tab 1095 \tab 8 \tab 5 \tab 27 \tab 12 \tab 1 \tab 9 \tab - \tab 10 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 647 \tab 33 \tab 8 \tab 150 \tab 11 \tab 1066 \tab 3 \tab 15 \tab 82 \tab 3 \tab 4 \tab 12 \tab - \tab 10 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 653 \tab 28 \tab 4 \tab 162 \tab 9 \tab 1015 \tab 8 \tab 3 \tab 50 \tab 15 \tab - \tab 13 \tab 4 \tab 10 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 872 \tab 41 \tab 12 \tab 206 \tab 11 \tab 1273 \tab 4 \tab 10 \tab 68 \tab 17 \tab - \tab 12 \tab 1 \tab 11 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 923 \tab 55 \tab 10 \tab 248 \tab 23 \tab 1511 \tab 10 \tab 12 \tab 30 \tab 13 \tab - \tab 20 \tab 2 \tab 10 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 732 \tab 42 \tab 7 \tab 241 \tab 10 \tab 1237 \tab 9 \tab 11 \tab 14 \tab 12 \tab - \tab 17 \tab 2 \tab 14 \tab - \tab - \tab 2348 \cr
   2021 \tab 40 \tab 1290 \tab 87 \tab 3 \tab 499 \tab 59 \tab 1860 \tab 29 \tab 32 \tab 60 \tab 30 \tab 6 \tab 1 \tab 2 \tab - \tab 2 \tab 32 \tab 4032 \cr
   2022 \tab 2 \tab 1159 \tab 55 \tab 18 \tab 404 \tab 27 \tab 1685 \tab 20 \tab 17 \tab 52 \tab 30 \tab 1 \tab 30 \tab - \tab 22 \tab 2 \tab 20 \tab 3544 \cr
   Total \tab 3173 \tab 20423 \tab 1408 \tab 270 \tab 4324 \tab 734 \tab 40587 \tab 107 \tab 177 \tab 529 \tab 174 \tab 38 \tab 192 \tab 26 \tab 160 \tab 16 \tab 52 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{6} buddhism
\item \code{7} hinduism
\item \code{8} other eastern religions
\item \code{9} muslim/islam
\item \code{10} orthodox-christian
\item \code{11} christian
\item \code{12} native american
\item \code{13} inter-nondenominational
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
