% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_35.R
\name{reliten}
\alias{reliten}
\title{reliten}
\source{
General Social Survey https://gss.norc.org
}
\description{
Strength of affiliation
}
\details{
Question 107. Would you call yoruself a strong \link{RELIGIOUS PREFERENCE} or a not very strong \link{RELIGIOUS PREFERENCE}?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab don't know \tab no answer \tab no religion \tab not very strong \tab somewhat strong (vol.) \tab strong \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 9 \tab 10 \tab 101 \tab 667 \tab 113 \tab 584 \tab - \tab 1484 \cr
   1975 \tab - \tab 3 \tab 16 \tab 113 \tab 626 \tab 152 \tab 580 \tab - \tab 1490 \cr
   1976 \tab - \tab 7 \tab 24 \tab 114 \tab 627 \tab 194 \tab 533 \tab - \tab 1499 \cr
   1977 \tab - \tab 7 \tab 23 \tab 93 \tab 715 \tab 111 \tab 581 \tab - \tab 1530 \cr
   1978 \tab - \tab 9 \tab 7 \tab 119 \tab 734 \tab 112 \tab 551 \tab - \tab 1532 \cr
   1980 \tab - \tab 7 \tab 16 \tab 105 \tab 649 \tab 133 \tab 558 \tab - \tab 1468 \cr
   1982 \tab - \tab 9 \tab 46 \tab 132 \tab 758 \tab 155 \tab 760 \tab - \tab 1860 \cr
   1983 \tab - \tab 6 \tab 19 \tab 117 \tab 707 \tab 106 \tab 644 \tab - \tab 1599 \cr
   1984 \tab - \tab 11 \tab 21 \tab 107 \tab 566 \tab 112 \tab 656 \tab - \tab 1473 \cr
   1985 \tab - \tab 7 \tab 46 \tab 109 \tab 630 \tab 135 \tab 607 \tab - \tab 1534 \cr
   1986 \tab - \tab 6 \tab 105 \tab 98 \tab 579 \tab 137 \tab 545 \tab - \tab 1470 \cr
   1987 \tab - \tab 12 \tab 29 \tab 121 \tab 720 \tab 186 \tab 751 \tab - \tab 1819 \cr
   1988 \tab - \tab 4 \tab 161 \tab 118 \tab 549 \tab 150 \tab 499 \tab - \tab 1481 \cr
   1989 \tab - \tab 13 \tab 75 \tab 120 \tab 600 \tab 207 \tab 522 \tab - \tab 1537 \cr
   1990 \tab - \tab 13 \tab 40 \tab 109 \tab 541 \tab 186 \tab 483 \tab - \tab 1372 \cr
   1991 \tab - \tab 15 \tab 48 \tab 102 \tab 598 \tab 214 \tab 540 \tab - \tab 1517 \cr
   1993 \tab - \tab 15 \tab 62 \tab 146 \tab 587 \tab 216 \tab 580 \tab - \tab 1606 \cr
   1994 \tab - \tab 22 \tab 92 \tab 274 \tab 1188 \tab 314 \tab 1102 \tab - \tab 2992 \cr
   1996 \tab - \tab 27 \tab 92 \tab 340 \tab 1126 \tab 301 \tab 1018 \tab - \tab 2904 \cr
   1998 \tab - \tab 28 \tab 138 \tab 401 \tab 958 \tab 285 \tab 1022 \tab - \tab 2832 \cr
   2000 \tab - \tab 36 \tab 200 \tab 399 \tab 957 \tab 287 \tab 938 \tab - \tab 2817 \cr
   2002 \tab - \tab 8 \tab 15 \tab 383 \tab 1041 \tab 308 \tab 1010 \tab - \tab 2765 \cr
   2004 \tab - \tab 8 \tab 36 \tab 403 \tab 1026 \tab 266 \tab 1073 \tab - \tab 2812 \cr
   2006 \tab - \tab 12 \tab 29 \tab 739 \tab 1632 \tab 482 \tab 1616 \tab - \tab 4510 \cr
   2008 \tab 1 \tab 3 \tab 78 \tab 332 \tab 744 \tab 144 \tab 721 \tab - \tab 2023 \cr
   2010 \tab - \tab 11 \tab 87 \tab 364 \tab 747 \tab 128 \tab 707 \tab - \tab 2044 \cr
   2012 \tab 5 \tab 6 \tab 8 \tab 387 \tab 690 \tab 165 \tab 713 \tab - \tab 1974 \cr
   2014 \tab 3 \tab 11 \tab 21 \tab 522 \tab 892 \tab 181 \tab 908 \tab - \tab 2538 \cr
   2016 \tab 3 \tab 4 \tab 23 \tab 619 \tab 1015 \tab 155 \tab 1048 \tab - \tab 2867 \cr
   2018 \tab 5 \tab 7 \tab 22 \tab 542 \tab 869 \tab 101 \tab 802 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 2188 \tab 22 \tab 3 \tab - \tab 560 \tab 171 \tab 600 \tab - \tab 3544 \cr
   Total \tab 5322 \tab 358 \tab 1592 \tab 7629 \tab 24298 \tab 5907 \tab 23252 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} strong
\item \code{2} not very strong
\item \code{3} somewhat strong (vol.)
\item \code{4} no religion
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
