% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_594.R
\name{relsp4}
\alias{relsp4}
\title{Relationship of person 4 to spouse of househol}
\source{
General Social Survey https://gss.norc.org
}
\description{
relsp4
}
\details{
Question 1633. What is (NAME)'s relationship to spouse?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab adopted child \tab brother, sister (full, half, or step sibling) \tab child of non-relative \tab child, not specified \tab cousin \tab friend \tab grandchild, great-grandchild \tab natural/biological child \tab niece or nephew \tab no answer \tab other non-relative (employee, boarder) \tab other relative (great-aunt, grandniece, etc.) \tab parent \tab parent-in-law \tab roommate, housemate \tab sibling-in-law \tab son-in-law, daughter-in-law \tab stepchild \tab uncle or aunt \tab don't know \tab grandparent, great-grandparent \tab partner, fiance(e), boyfriend, girlfriend, etc. \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2241 \tab 13 \tab 3 \tab 7 \tab 56 \tab 1 \tab 3 \tab 18 \tab 373 \tab 2 \tab 65 \tab 4 \tab 1 \tab 6 \tab 2 \tab 1 \tab 3 \tab 6 \tab 12 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2370 \tab 5 \tab 4 \tab 4 \tab 76 \tab - \tab 1 \tab 12 \tab 259 \tab 1 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 5 \tab 13 \tab 1 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2305 \tab 1 \tab 2 \tab 7 \tab 150 \tab - \tab 2 \tab 13 \tab 290 \tab 3 \tab - \tab 5 \tab - \tab 2 \tab 1 \tab 4 \tab 3 \tab 4 \tab 17 \tab - \tab 2 \tab 1 \tab - \tab - \tab 2812 \cr
   2006 \tab 3624 \tab 6 \tab 5 \tab 6 \tab 297 \tab 2 \tab 6 \tab 18 \tab 479 \tab 1 \tab - \tab 13 \tab 1 \tab 4 \tab 4 \tab 6 \tab 7 \tab 3 \tab 26 \tab 1 \tab 1 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1634 \tab - \tab - \tab 8 \tab 180 \tab - \tab 2 \tab 15 \tab 155 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 4 \tab 4 \tab 11 \tab - \tab 1 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1714 \tab 2 \tab 3 \tab 3 \tab 167 \tab - \tab 2 \tab 11 \tab 102 \tab 4 \tab 2 \tab 5 \tab 1 \tab 3 \tab 1 \tab 5 \tab 2 \tab 10 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1603 \tab 3 \tab 5 \tab 5 \tab 178 \tab - \tab 2 \tab 13 \tab 125 \tab 3 \tab - \tab 6 \tab - \tab 1 \tab 3 \tab 8 \tab 1 \tab 4 \tab 12 \tab - \tab 2 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2120 \tab 1 \tab - \tab 5 \tab 216 \tab - \tab 4 \tab 9 \tab 141 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab 10 \tab 13 \tab 2 \tab 1 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2448 \tab 6 \tab 2 \tab 5 \tab 276 \tab - \tab 3 \tab 10 \tab 70 \tab 4 \tab - \tab 6 \tab - \tab 2 \tab 4 \tab 4 \tab 1 \tab 7 \tab 17 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2867 \cr
   2018 \tab 1996 \tab 1 \tab 2 \tab 4 \tab 187 \tab 1 \tab 2 \tab 11 \tab 99 \tab 2 \tab 1 \tab - \tab - \tab 4 \tab - \tab 5 \tab 3 \tab 14 \tab 15 \tab - \tab - \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 63715 \tab 38 \tab 26 \tab 54 \tab 1783 \tab 4 \tab 27 \tab 130 \tab 2093 \tab 23 \tab 74 \tab 47 \tab 5 \tab 29 \tab 19 \tab 40 \tab 26 \tab 67 \tab 142 \tab 5 \tab 8 \tab 2 \tab 1 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} hh spouse, partner
\item \code{2} spouse
\item \code{3} partner, fiance(e), boyfriend, girlfriend, etc.
\item \code{4} child, not specified
\item \code{5} grandchild, great-grandchild
\item \code{6} son-in-law, daughter-in-law
\item \code{7} parent
\item \code{8} parent-in-law
\item \code{9} grandparent, great-grandparent
\item \code{10} uncle or aunt
\item \code{11} niece or nephew
\item \code{12} cousin
\item \code{13} brother, sister (full, half, or step sibling)
\item \code{14} sibling-in-law
\item \code{15} other relative (great-aunt, grandniece, etc.)
\item \code{16} roommate, housemate
\item \code{17} friend
\item \code{18} child of non-relative
\item \code{19} other non-relative (employee, boarder)
\item \code{41} natural/biological child
\item \code{42} adopted child
\item \code{43} stepchild
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
