% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_76.R
\name{rifle}
\alias{rifle}
\title{rifle}
\source{
General Social Survey https://gss.norc.org
}
\description{
Rifle in home
}
\details{
Question 237a. If yes: Is it a pistol, shotgun, rigfle, or what? RIFLE
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab no answer \tab refused \tab yes \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 1045 \tab 7 \tab 15 \tab 437 \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 1073 \tab 5 \tab 10 \tab 396 \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 1052 \tab 10 \tab 17 \tab 420 \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 1057 \tab 11 \tab 2 \tab 460 \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 1025 \tab 10 \tab 7 \tab 426 \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 1373 \tab 6 \tab 26 \tab 455 \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 1058 \tab 6 \tab 9 \tab 400 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 1087 \tab 4 \tab 10 \tab 433 \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 1360 \tab 7 \tab 10 \tab 442 \tab - \tab - \tab 1819 \cr
   1988 \tab 504 \tab 725 \tab 7 \tab 10 \tab 235 \tab - \tab - \tab 1481 \cr
   1989 \tab 504 \tab 746 \tab 4 \tab - \tab 283 \tab - \tab - \tab 1537 \cr
   1990 \tab 455 \tab 677 \tab 9 \tab - \tab 230 \tab 1 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 730 \tab 18 \tab - \tab 245 \tab - \tab - \tab 1517 \cr
   1993 \tab 531 \tab 817 \tab 2 \tab 7 \tab 249 \tab - \tab - \tab 1606 \cr
   1994 \tab 996 \tab 1485 \tab 23 \tab 5 \tab 483 \tab - \tab - \tab 2992 \cr
   1996 \tab 981 \tab 1464 \tab 5 \tab 7 \tab 447 \tab - \tab - \tab 2904 \cr
   1998 \tab 951 \tab 1476 \tab 9 \tab 6 \tab 390 \tab - \tab - \tab 2832 \cr
   2000 \tab 956 \tab 1459 \tab 14 \tab 23 \tab 364 \tab 1 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 731 \tab - \tab 10 \tab 183 \tab - \tab - \tab 2765 \cr
   2004 \tab 1916 \tab 695 \tab 11 \tab 16 \tab 174 \tab - \tab - \tab 2812 \cr
   2006 \tab 2509 \tab 1582 \tab - \tab 35 \tab 381 \tab 3 \tab - \tab 4510 \cr
   2008 \tab 674 \tab 1050 \tab - \tab 19 \tab 276 \tab 4 \tab - \tab 2023 \cr
   2010 \tab 766 \tab 983 \tab - \tab 50 \tab 240 \tab 5 \tab - \tab 2044 \cr
   2012 \tab 669 \tab 1010 \tab - \tab 28 \tab 260 \tab 7 \tab - \tab 1974 \cr
   2014 \tab 827 \tab 1325 \tab - \tab 63 \tab 317 \tab 6 \tab - \tab 2538 \cr
   2016 \tab 978 \tab 1458 \tab - \tab 60 \tab 365 \tab 6 \tab - \tab 2867 \cr
   2018 \tab 779 \tab 1197 \tab 4 \tab 39 \tab 325 \tab 4 \tab - \tab 2348 \cr
   2021 \tab - \tab 3105 \tab 4 \tab 117 \tab 721 \tab 76 \tab 9 \tab 4032 \cr
   2022 \tab 1203 \tab 1816 \tab 6 \tab 72 \tab 416 \tab 24 \tab 7 \tab 3544 \cr
   Total \tab 26268 \tab 34661 \tab 182 \tab 673 \tab 10453 \tab 137 \tab 16 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} refused
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
