% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_17.R
\name{rincom06}
\alias{rincom06}
\title{rincom06}
\source{
General Social Survey https://gss.norc.org
}
\description{
Respondents income
}
\details{
Question In which of these groups did your earnings from \link{OCCUPATION} from all sources for \link{last year} fall? That is, before taxes or other deductions.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllll}{
   year \tab iap \tab $1 000 to 2 999 \tab $10000 to 12499 \tab $110000 to $129999 \tab $12500 to 14999 \tab $130000 to $149999 \tab $15000 to 17499 \tab $150000 or over \tab $17500 to 19999 \tab $20000 to 22499 \tab $22500 to 24999 \tab $25000 to 29999 \tab $3 000 to 3 999 \tab $30000 to 34999 \tab $35000 to 39999 \tab $4 000 to 4 999 \tab $40000 to 49999 \tab $5 000 to 5 999 \tab $50000 to 59999 \tab $6 000 to 6 999 \tab $60000 to 74999 \tab $7 000 to 7 999 \tab $75000 to $89999 \tab $8 000 to 9 999 \tab $90000 to $109999 \tab don't know \tab no answer \tab refused \tab under $1 000 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 1472 \tab 75 \tab 123 \tab 50 \tab 108 \tab 22 \tab 110 \tab 61 \tab 111 \tab 139 \tab 109 \tab 193 \tab 47 \tab 222 \tab 169 \tab 53 \tab 301 \tab 41 \tab 198 \tab 42 \tab 176 \tab 37 \tab 104 \tab 61 \tab 64 \tab 71 \tab 53 \tab 245 \tab 53 \tab 4510 \cr
   2008 \tab 693 \tab 39 \tab 47 \tab 28 \tab 63 \tab 6 \tab 40 \tab 32 \tab 42 \tab 44 \tab 76 \tab 70 \tab 27 \tab 94 \tab 71 \tab 13 \tab 125 \tab 19 \tab 85 \tab 16 \tab 87 \tab 20 \tab 61 \tab 24 \tab 38 \tab 16 \tab 40 \tab 85 \tab 22 \tab 2023 \cr
   2010 \tab 684 \tab 43 \tab 72 \tab 28 \tab 38 \tab 14 \tab 51 \tab 22 \tab 37 \tab 56 \tab 51 \tab 66 \tab 33 \tab 78 \tab 67 \tab 23 \tab 102 \tab 26 \tab 88 \tab 28 \tab 98 \tab 19 \tab 63 \tab 34 \tab 36 \tab 13 \tab - \tab 145 \tab 29 \tab 2044 \cr
   2012 \tab 723 \tab 32 \tab 54 \tab 21 \tab 49 \tab 7 \tab 44 \tab 35 \tab 45 \tab 51 \tab 54 \tab 78 \tab 30 \tab 71 \tab 62 \tab 22 \tab 121 \tab 20 \tab 78 \tab 24 \tab 84 \tab 22 \tab 51 \tab 23 \tab 41 \tab 28 \tab - \tab 77 \tab 27 \tab 1974 \cr
   2014 \tab 923 \tab 36 \tab 56 \tab 37 \tab 55 \tab 21 \tab 48 \tab 58 \tab 39 \tab 76 \tab 67 \tab 90 \tab 38 \tab 112 \tab 79 \tab 25 \tab 154 \tab 32 \tab 145 \tab 24 \tab 114 \tab 14 \tab 88 \tab 30 \tab 53 \tab 19 \tab - \tab 73 \tab 32 \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 63796 \tab 225 \tab 352 \tab 164 \tab 313 \tab 70 \tab 293 \tab 208 \tab 274 \tab 366 \tab 357 \tab 497 \tab 175 \tab 577 \tab 448 \tab 136 \tab 803 \tab 138 \tab 594 \tab 134 \tab 559 \tab 112 \tab 367 \tab 172 \tab 232 \tab 147 \tab 93 \tab 625 \tab 163 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} under $1 000
\item \code{2} $1 000 to 2 999
\item \code{3} $3 000 to 3 999
\item \code{4} $4 000 to 4 999
\item \code{5} $5 000 to 5 999
\item \code{6} $6 000 to 6 999
\item \code{7} $7 000 to 7 999
\item \code{8} $8 000 to 9 999
\item \code{9} $10000 to 12499
\item \code{10} $12500 to 14999
\item \code{11} $15000 to 17499
\item \code{12} $17500 to 19999
\item \code{13} $20000 to 22499
\item \code{14} $22500 to 24999
\item \code{15} $25000 to 29999
\item \code{16} $30000 to 34999
\item \code{17} $35000 to 39999
\item \code{18} $40000 to 49999
\item \code{19} $50000 to 59999
\item \code{20} $60000 to 74999
\item \code{21} $75000 to $89999
\item \code{22} $90000 to $109999
\item \code{23} $110000 to $129999
\item \code{24} $130000 to $149999
\item \code{25} $150000 or over
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
