% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_17.R
\name{rincom16}
\alias{rincom16}
\title{Respondents income}
\source{
General Social Survey https://gss.norc.org
}
\description{
rincom16
}
\details{
Question In which of these groups did your earnings from (OCC), from all sources for (last year) fall? That is, before taxes or other deductions. Just tell me the letter.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllll}{
   year \tab iap \tab $1,000 to $2,999 \tab $10,000 to $12,499 \tab $110,000 to $129,999 \tab $12,500 to $14,999 \tab $130,000 to $149,999 \tab $15,000 to $17,499 \tab $150,000 to $169,999 \tab $17,500 to $19,999 \tab $170,000 or over \tab $20,000 to $22,499 \tab $22,500 to $24,999 \tab $25,000 to $29,999 \tab $3,000 to $3,999 \tab $30,000 to $34,999 \tab $35,000 to $39,999 \tab $4,000 to $4,999 \tab $40,000 to $49,999 \tab $5,000 to $5,999 \tab $50,000 to $59,999 \tab $6,000 to $6,999 \tab $60,000 to $74,999 \tab $7,000 to $7,999 \tab $75,000 to $89,999 \tab $8,000 to $9,999 \tab $90,000 to $109,999 \tab don't know \tab refused \tab under $1,000 \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1136 \tab 51 \tab 51 \tab 36 \tab 45 \tab 21 \tab 60 \tab 14 \tab 52 \tab 51 \tab 74 \tab 64 \tab 97 \tab 32 \tab 118 \tab 108 \tab 30 \tab 158 \tab 31 \tab 137 \tab 31 \tab 141 \tab 24 \tab 79 \tab 34 \tab 68 \tab 14 \tab 85 \tab 25 \tab - \tab 2867 \cr
   2018 \tab 899 \tab 32 \tab 48 \tab 43 \tab 46 \tab 32 \tab 25 \tab 14 \tab 36 \tab 48 \tab 61 \tab 46 \tab 73 \tab 32 \tab 96 \tab 81 \tab 21 \tab 136 \tab 21 \tab 114 \tab 12 \tab 114 \tab 18 \tab 81 \tab 33 \tab 67 \tab 26 \tab 60 \tab 33 \tab - \tab 2348 \cr
   2021 \tab 1307 \tab 52 \tab 88 \tab 96 \tab 61 \tab 54 \tab 57 \tab 49 \tab 62 \tab 129 \tab 73 \tab 74 \tab 124 \tab 28 \tab 148 \tab 122 \tab 24 \tab 220 \tab 32 \tab 203 \tab 27 \tab 267 \tab 30 \tab 191 \tab 43 \tab 166 \tab 161 \tab 63 \tab 36 \tab 45 \tab 4032 \cr
   2022 \tab 1308 \tab 35 \tab 70 \tab 79 \tab 43 \tab 56 \tab 49 \tab 32 \tab 42 \tab 115 \tab 78 \tab 62 \tab 93 \tab 29 \tab 123 \tab 108 \tab 20 \tab 161 \tab 26 \tab 188 \tab 17 \tab 205 \tab 20 \tab 146 \tab 52 \tab 118 \tab 111 \tab 113 \tab 23 \tab 22 \tab 3544 \cr
   Total \tab 64249 \tab 170 \tab 257 \tab 254 \tab 195 \tab 163 \tab 191 \tab 109 \tab 192 \tab 343 \tab 286 \tab 246 \tab 387 \tab 121 \tab 485 \tab 419 \tab 95 \tab 675 \tab 110 \tab 642 \tab 87 \tab 727 \tab 92 \tab 497 \tab 162 \tab 419 \tab 312 \tab 321 \tab 117 \tab 67 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} under $1,000
\item \code{2} $1,000 to $2,999
\item \code{3} $3,000 to $3,999
\item \code{4} $4,000 to $4,999
\item \code{5} $5,000 to $5,999
\item \code{6} $6,000 to $6,999
\item \code{7} $7,000 to $7,999
\item \code{8} $8,000 to $9,999
\item \code{9} $10,000 to $12,499
\item \code{10} $12,500 to $14,999
\item \code{11} $15,000 to $17,499
\item \code{12} $17,500 to $19,999
\item \code{13} $20,000 to $22,499
\item \code{14} $22,500 to $24,999
\item \code{15} $25,000 to $29,999
\item \code{16} $30,000 to $34,999
\item \code{17} $35,000 to $39,999
\item \code{18} $40,000 to $49,999
\item \code{19} $50,000 to $59,999
\item \code{20} $60,000 to $74,999
\item \code{21} $75,000 to $89,999
\item \code{22} $90,000 to $109,999
\item \code{23} $110,000 to $129,999
\item \code{24} $130,000 to $149,999
\item \code{25} $150,000 to $169,999
\item \code{26} $170,000 or over
\item \code{27} refused
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
