% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_592.R
\name{rvisitor}
\alias{rvisitor}
\title{Is r a visitor?}
\source{
General Social Survey https://gss.norc.org
}
\description{
rvisitor
}
\details{
Question 1630. Is respondent a visitor?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab r is household member \tab r is visitor \tab no answer \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 1488 \tab 2 \tab - \tab - \tab 1490 \cr
   1976 \tab 755 \tab 742 \tab 2 \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 1526 \tab 4 \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 1527 \tab 5 \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 1467 \tab 1 \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 1857 \tab 3 \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 1594 \tab 2 \tab 3 \tab - \tab 1599 \cr
   1984 \tab - \tab 1470 \tab 3 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 1530 \tab 4 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 1470 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 1813 \tab 6 \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 1480 \tab 1 \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 1531 \tab 6 \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 1371 \tab 1 \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 1513 \tab 4 \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 1605 \tab 1 \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 2988 \tab 4 \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 2900 \tab 3 \tab 1 \tab - \tab 2904 \cr
   1998 \tab - \tab 2830 \tab 2 \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 2814 \tab 2 \tab 1 \tab - \tab 2817 \cr
   2002 \tab - \tab 2763 \tab 2 \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 2808 \tab 4 \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 4499 \tab 11 \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 2019 \tab 4 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 2038 \tab 6 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1967 \tab 7 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2534 \tab 4 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2857 \tab 10 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 2342 \tab 6 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 8900 \tab 59343 \tab 110 \tab 5 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} r is household member
\item \code{2} r is visitor
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
