% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{sample}
\alias{sample}
\title{Sampling frame and method}
\source{
General Social Survey https://gss.norc.org
}
\description{
sample
}
\details{
Question 1664. The GSS has employed various sampling procedures and frames.
}
\section{Values}{

\itemize{
\item \code{1} 1960 sample frame, block quota
\item \code{2} 1970 sample frame, block quota
\item \code{3} 1970 sample frame, full probability
\item \code{4} 1970 sample frame, full probability black oversample
\item \code{5} 1980 black sample frame, full probability black oversample
\item \code{6} 1980 sample frame, full probability
\item \code{7} 1980 sample frame, full probability black oversample
\item \code{8} 1990 sample frame, full probability
\item \code{9} 2000 sample frame, full probability
\item \code{10} 2010 sample frame, full probability
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
