% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_50.R
\name{satcity}
\alias{satcity}
\title{satcity}
\source{
General Social Survey https://gss.norc.org
}
\description{
City or place r lives in
}
\details{
Question 164. For each area of life I am going to name, tell me the number that shows how much satisfaction you get from that area. a. The city or place you live in.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab a fair amount \tab a little \tab don't know \tab great deal \tab none \tab quite a bit \tab some \tab very great deal \tab no answer \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 337 \tab 77 \tab 2 \tab 357 \tab 40 \tab 254 \tab 91 \tab 346 \tab - \tab 1504 \cr
   1974 \tab - \tab 343 \tab 59 \tab 1 \tab 394 \tab 30 \tab 257 \tab 103 \tab 297 \tab - \tab 1484 \cr
   1975 \tab - \tab 298 \tab 80 \tab 2 \tab 438 \tab 41 \tab 222 \tab 93 \tab 311 \tab 5 \tab 1490 \cr
   1976 \tab - \tab 307 \tab 57 \tab 4 \tab 446 \tab 29 \tab 260 \tab 101 \tab 293 \tab 2 \tab 1499 \cr
   1977 \tab - \tab 330 \tab 85 \tab 3 \tab 421 \tab 37 \tab 262 \tab 95 \tab 295 \tab 2 \tab 1530 \cr
   1978 \tab - \tab 305 \tab 71 \tab 1 \tab 460 \tab 43 \tab 291 \tab 101 \tab 254 \tab 6 \tab 1532 \cr
   1980 \tab - \tab 267 \tab 65 \tab 6 \tab 450 \tab 27 \tab 224 \tab 82 \tab 347 \tab - \tab 1468 \cr
   1982 \tab - \tab 446 \tab 96 \tab 5 \tab 442 \tab 44 \tab 358 \tab 161 \tab 304 \tab 4 \tab 1860 \cr
   1983 \tab - \tab 337 \tab 85 \tab 6 \tab 488 \tab 28 \tab 280 \tab 119 \tab 253 \tab 3 \tab 1599 \cr
   1984 \tab - \tab 262 \tab 72 \tab 6 \tab 478 \tab 20 \tab 256 \tab 85 \tab 288 \tab 6 \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 315 \tab 92 \tab 3 \tab 395 \tab 28 \tab 274 \tab 117 \tab 232 \tab 14 \tab 1470 \cr
   1987 \tab - \tab 383 \tab 81 \tab 1 \tab 557 \tab 32 \tab 304 \tab 130 \tab 320 \tab 11 \tab 1819 \cr
   1988 \tab 484 \tab 206 \tab 36 \tab 1 \tab 307 \tab 23 \tab 187 \tab 58 \tab 176 \tab 3 \tab 1481 \cr
   1989 \tab 502 \tab 196 \tab 63 \tab 2 \tab 283 \tab 22 \tab 217 \tab 69 \tab 178 \tab 5 \tab 1537 \cr
   1990 \tab 473 \tab 183 \tab 38 \tab - \tab 299 \tab 12 \tab 161 \tab 69 \tab 136 \tab 1 \tab 1372 \cr
   1991 \tab 500 \tab 207 \tab 44 \tab - \tab 302 \tab 11 \tab 187 \tab 88 \tab 176 \tab 2 \tab 1517 \cr
   1993 \tab 549 \tab 211 \tab 57 \tab - \tab 304 \tab 21 \tab 195 \tab 84 \tab 182 \tab 3 \tab 1606 \cr
   1994 \tab 2481 \tab 99 \tab 16 \tab - \tab 161 \tab 16 \tab 105 \tab 33 \tab 80 \tab 1 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 48146 \tab 5032 \tab 1174 \tab 43 \tab 6982 \tab 504 \tab 4294 \tab 1679 \tab 4468 \tab 68 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} very great deal
\item \code{2} great deal
\item \code{3} quite a bit
\item \code{4} a fair amount
\item \code{5} some
\item \code{6} a little
\item \code{7} none
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
