% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_50.R
\name{satfam}
\alias{satfam}
\title{satfam}
\source{
General Social Survey https://gss.norc.org
}
\description{
Family life
}
\details{
Question 164. For each area of life I am going to name, tell me the number that shows how much satisfaction you get from that area. c. Your family life.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab a fair amount \tab a little \tab don't know \tab great deal \tab no answer \tab none \tab quite a bit \tab some \tab very great deal \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 128 \tab 33 \tab 5 \tab 468 \tab 6 \tab 25 \tab 146 \tab 51 \tab 642 \tab 1504 \cr
   1974 \tab - \tab 102 \tab 30 \tab 3 \tab 496 \tab 1 \tab 19 \tab 157 \tab 35 \tab 641 \tab 1484 \cr
   1975 \tab - \tab 98 \tab 23 \tab 4 \tab 488 \tab 4 \tab 28 \tab 155 \tab 37 \tab 653 \tab 1490 \cr
   1976 \tab - \tab 92 \tab 33 \tab 3 \tab 564 \tab 6 \tab 31 \tab 161 \tab 39 \tab 570 \tab 1499 \cr
   1977 \tab - \tab 112 \tab 21 \tab 6 \tab 501 \tab 3 \tab 26 \tab 181 \tab 43 \tab 637 \tab 1530 \cr
   1978 \tab - \tab 113 \tab 33 \tab 5 \tab 543 \tab 11 \tab 24 \tab 164 \tab 48 \tab 591 \tab 1532 \cr
   1980 \tab - \tab 90 \tab 27 \tab 7 \tab 496 \tab 2 \tab 23 \tab 148 \tab 27 \tab 648 \tab 1468 \cr
   1982 \tab - \tab 123 \tab 48 \tab 8 \tab 502 \tab 5 \tab 38 \tab 212 \tab 60 \tab 864 \tab 1860 \cr
   1983 \tab - \tab 112 \tab 27 \tab 11 \tab 559 \tab 6 \tab 39 \tab 182 \tab 51 \tab 612 \tab 1599 \cr
   1984 \tab - \tab 93 \tab 35 \tab 2 \tab 460 \tab 6 \tab 21 \tab 159 \tab 34 \tab 663 \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 124 \tab 34 \tab 5 \tab 501 \tab 17 \tab 28 \tab 180 \tab 48 \tab 533 \tab 1470 \cr
   1987 \tab 1 \tab 141 \tab 44 \tab 4 \tab 595 \tab 13 \tab 29 \tab 209 \tab 57 \tab 726 \tab 1819 \cr
   1988 \tab 484 \tab 60 \tab 28 \tab 1 \tab 335 \tab 3 \tab 20 \tab 86 \tab 30 \tab 434 \tab 1481 \cr
   1989 \tab 502 \tab 67 \tab 22 \tab - \tab 344 \tab 7 \tab 16 \tab 102 \tab 44 \tab 433 \tab 1537 \cr
   1990 \tab 473 \tab 63 \tab 17 \tab 1 \tab 305 \tab 1 \tab 17 \tab 94 \tab 34 \tab 367 \tab 1372 \cr
   1991 \tab 500 \tab 90 \tab 26 \tab 4 \tab 335 \tab 1 \tab 17 \tab 100 \tab 27 \tab 417 \tab 1517 \cr
   1993 \tab 549 \tab 78 \tab 26 \tab 4 \tab 361 \tab 6 \tab 20 \tab 99 \tab 31 \tab 432 \tab 1606 \cr
   1994 \tab 2481 \tab 35 \tab 13 \tab - \tab 179 \tab 2 \tab 10 \tab 53 \tab 17 \tab 202 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 48147 \tab 1721 \tab 520 \tab 73 \tab 8032 \tab 100 \tab 431 \tab 2588 \tab 713 \tab 10065 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} very great deal
\item \code{2} great deal
\item \code{3} quite a bit
\item \code{4} a fair amount
\item \code{5} some
\item \code{6} a little
\item \code{7} none
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
