% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_51.R
\name{satfrnd}
\alias{satfrnd}
\title{satfrnd}
\source{
General Social Survey https://gss.norc.org
}
\description{
Friendships
}
\details{
Question 164. For each area of life I am going to name, tell me the number that shows how much satisfaction you get from that area. d. Your friendships.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab a fair amount \tab a little \tab don't know \tab great deal \tab no answer \tab none \tab quite a bit \tab some \tab very great deal \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 158 \tab 37 \tab 3 \tab 547 \tab 6 \tab 9 \tab 204 \tab 43 \tab 497 \tab 1504 \cr
   1974 \tab - \tab 120 \tab 27 \tab - \tab 602 \tab - \tab 8 \tab 212 \tab 44 \tab 471 \tab 1484 \cr
   1975 \tab - \tab 129 \tab 33 \tab 1 \tab 623 \tab 5 \tab 8 \tab 217 \tab 47 \tab 427 \tab 1490 \cr
   1976 \tab - \tab 142 \tab 26 \tab 4 \tab 597 \tab 3 \tab 8 \tab 241 \tab 40 \tab 438 \tab 1499 \cr
   1977 \tab - \tab 145 \tab 31 \tab 1 \tab 595 \tab 4 \tab 10 \tab 234 \tab 45 \tab 465 \tab 1530 \cr
   1978 \tab - \tab 154 \tab 23 \tab - \tab 624 \tab 6 \tab 6 \tab 257 \tab 50 \tab 412 \tab 1532 \cr
   1980 \tab - \tab 112 \tab 18 \tab 1 \tab 609 \tab - \tab 12 \tab 184 \tab 30 \tab 502 \tab 1468 \cr
   1982 \tab - \tab 222 \tab 42 \tab 5 \tab 692 \tab 4 \tab 20 \tab 329 \tab 68 \tab 478 \tab 1860 \cr
   1983 \tab - \tab 147 \tab 29 \tab 7 \tab 646 \tab 3 \tab 13 \tab 282 \tab 48 \tab 424 \tab 1599 \cr
   1984 \tab - \tab 134 \tab 26 \tab 3 \tab 558 \tab 8 \tab 4 \tab 179 \tab 27 \tab 534 \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 160 \tab 29 \tab 4 \tab 607 \tab 15 \tab 13 \tab 244 \tab 51 \tab 347 \tab 1470 \cr
   1987 \tab - \tab 173 \tab 41 \tab 2 \tab 687 \tab 11 \tab 15 \tab 272 \tab 60 \tab 558 \tab 1819 \cr
   1988 \tab 484 \tab 83 \tab 24 \tab 2 \tab 408 \tab 2 \tab 5 \tab 154 \tab 34 \tab 285 \tab 1481 \cr
   1989 \tab 502 \tab 78 \tab 14 \tab 1 \tab 397 \tab 5 \tab 8 \tab 156 \tab 32 \tab 344 \tab 1537 \cr
   1990 \tab 473 \tab 62 \tab 22 \tab 2 \tab 373 \tab 2 \tab 9 \tab 117 \tab 30 \tab 282 \tab 1372 \cr
   1991 \tab 500 \tab 104 \tab 30 \tab - \tab 384 \tab 1 \tab 9 \tab 142 \tab 31 \tab 316 \tab 1517 \cr
   1993 \tab 549 \tab 108 \tab 27 \tab - \tab 387 \tab 4 \tab 5 \tab 151 \tab 33 \tab 342 \tab 1606 \cr
   1994 \tab 2481 \tab 46 \tab 3 \tab - \tab 195 \tab 1 \tab 5 \tab 87 \tab 16 \tab 158 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 48146 \tab 2277 \tab 482 \tab 36 \tab 9531 \tab 80 \tab 167 \tab 3662 \tab 729 \tab 7280 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} very great deal
\item \code{2} great deal
\item \code{3} quite a bit
\item \code{4} a fair amount
\item \code{5} some
\item \code{6} a little
\item \code{7} none
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
