% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_61.R
\name{satjob}
\alias{satjob}
\title{satjob}
\source{
General Social Survey https://gss.norc.org
}
\description{
Work satisfaction
}
\details{
Question 181. On the whole, how satisfied are you with the work you do?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab a little dissatisfied \tab don't know \tab iap \tab moderately satisfied \tab no answer \tab very dissatisfied \tab very satisfied \tab skipped on web \tab Total \cr
   1972 \tab 106 \tab 3 \tab 660 \tab 347 \tab 6 \tab 33 \tab 458 \tab - \tab 1613 \cr
   1973 \tab 92 \tab 2 \tab 231 \tab 436 \tab 130 \tab 49 \tab 564 \tab - \tab 1504 \cr
   1974 \tab 122 \tab 5 \tab 224 \tab 458 \tab 32 \tab 57 \tab 586 \tab - \tab 1484 \cr
   1975 \tab 105 \tab 3 \tab 295 \tab 381 \tab 27 \tab 48 \tab 631 \tab - \tab 1490 \cr
   1976 \tab 108 \tab 3 \tab 294 \tab 407 \tab 17 \tab 53 \tab 617 \tab - \tab 1499 \cr
   1977 \tab 130 \tab 1 \tab 213 \tab 496 \tab 54 \tab 35 \tab 601 \tab - \tab 1530 \cr
   1978 \tab 108 \tab 1 \tab 222 \tab 463 \tab 29 \tab 58 \tab 651 \tab - \tab 1532 \cr
   1980 \tab 159 \tab 1 \tab 214 \tab 444 \tab 7 \tab 59 \tab 584 \tab - \tab 1468 \cr
   1982 \tab 160 \tab 3 \tab 301 \tab 562 \tab 64 \tab 84 \tab 686 \tab - \tab 1860 \cr
   1983 \tab 117 \tab - \tab 221 \tab 499 \tab 45 \tab 60 \tab 657 \tab - \tab 1599 \cr
   1984 \tab 146 \tab - \tab 208 \tab 423 \tab 57 \tab 84 \tab 555 \tab - \tab 1473 \cr
   1985 \tab 122 \tab 2 \tab 263 \tab 467 \tab 34 \tab 55 \tab 591 \tab - \tab 1534 \cr
   1986 \tab 104 \tab 1 \tab 257 \tab 458 \tab 50 \tab 30 \tab 570 \tab - \tab 1470 \cr
   1987 \tab 186 \tab 1 \tab 324 \tab 559 \tab 55 \tab 77 \tab 617 \tab - \tab 1819 \cr
   1988 \tab 117 \tab 2 \tab 290 \tab 455 \tab 36 \tab 44 \tab 537 \tab - \tab 1481 \cr
   1989 \tab 125 \tab 2 \tab 287 \tab 472 \tab 42 \tab 52 \tab 557 \tab - \tab 1537 \cr
   1990 \tab 109 \tab - \tab 263 \tab 413 \tab 68 \tab 38 \tab 481 \tab - \tab 1372 \cr
   1991 \tab 108 \tab 3 \tab 296 \tab 482 \tab 69 \tab 49 \tab 510 \tab - \tab 1517 \cr
   1993 \tab 133 \tab 1 \tab 329 \tab 505 \tab 48 \tab 58 \tab 532 \tab - \tab 1606 \cr
   1994 \tab 247 \tab 7 \tab 551 \tab 932 \tab 96 \tab 81 \tab 1078 \tab - \tab 2992 \cr
   1996 \tab 253 \tab 3 \tab 488 \tab 918 \tab 100 \tab 101 \tab 1041 \tab - \tab 2904 \cr
   1998 \tab 223 \tab 7 \tab 527 \tab 851 \tab 82 \tab 75 \tab 1067 \tab - \tab 2832 \cr
   2000 \tab 186 \tab 8 \tab 542 \tab 918 \tab 105 \tab 77 \tab 981 \tab - \tab 2817 \cr
   2002 \tab 100 \tab 13 \tab 1689 \tab 386 \tab 2 \tab 41 \tab 534 \tab - \tab 2765 \cr
   2004 \tab 125 \tab 16 \tab 1375 \tab 507 \tab 24 \tab 63 \tab 702 \tab - \tab 2812 \cr
   2006 \tab 192 \tab 36 \tab 2237 \tab 823 \tab 60 \tab 89 \tab 1073 \tab - \tab 4510 \cr
   2008 \tab 135 \tab 19 \tab 453 \tab 570 \tab 15 \tab 53 \tab 778 \tab - \tab 2023 \cr
   2010 \tab 155 \tab 20 \tab 477 \tab 565 \tab 19 \tab 62 \tab 746 \tab - \tab 2044 \cr
   2012 \tab 136 \tab 11 \tab 481 \tab 539 \tab 25 \tab 50 \tab 732 \tab - \tab 1974 \cr
   2014 \tab 161 \tab 12 \tab 626 \tab 686 \tab 5 \tab 92 \tab 956 \tab - \tab 2538 \cr
   2016 \tab 167 \tab 15 \tab 1064 \tab 645 \tab 9 \tab 67 \tab 900 \tab - \tab 2867 \cr
   2018 \tab 174 \tab 8 \tab 574 \tab 656 \tab 27 \tab 65 \tab 844 \tab - \tab 2348 \cr
   2021 \tab 305 \tab 5 \tab 1199 \tab 1119 \tab 60 \tab 112 \tab 1198 \tab 34 \tab 4032 \cr
   2022 \tab 241 \tab 7 \tab 1053 \tab 1012 \tab 8 \tab 107 \tab 1103 \tab 13 \tab 3544 \cr
   Total \tab 5157 \tab 221 \tab 18728 \tab 19854 \tab 1507 \tab 2158 \tab 24718 \tab 47 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} very satisfied
\item \code{2} moderately satisfied
\item \code{3} a little dissatisfied
\item \code{4} very dissatisfied
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
