% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_169.R
\name{sbocc10}
\alias{sbocc10}
\title{sbocc10}
\source{
General Social Survey https://gss.norc.org
}
\description{
R sibling census occupation code (2010)
}
\details{
Question None
}
\section{Values}{

\itemize{
\item \code{10} chief executives
\item \code{20} general and operations managers
\item \code{30} legislators
\item \code{40} advertising and promotions managers
\item \code{50} marketing and sales managers
\item \code{60} public relations and fundraising managers
\item \code{100} administrative services managers
\item \code{110} computer and information systems managers
\item \code{120} financial managers
\item \code{135} compensation and benefits managers
\item \code{136} human resources managers
\item \code{137} training and development managers
\item \code{140} industrial production managers
\item \code{150} purchasing managers
\item \code{160} transportation, storage, and distribution managers
\item \code{205} farmers, ranchers, and other agricultural managers
\item \code{220} construction managers
\item \code{230} education administrators
\item \code{300} architectural and engineering managers
\item \code{310} food service managers
\item \code{325} funeral service managers
\item \code{330} gaming managers
\item \code{340} lodging managers
\item \code{350} medical and health services managers
\item \code{360} natural sciences managers
\item \code{400} postmasters and mail superintendents
\item \code{410} property, real estate, and community association managers
\item \code{420} social and community service managers
\item \code{425} emergency management directors
\item \code{430} managers, all other
\item \code{500} agents and business managers of artists, performers,
\item \code{510} buyers and purchasing agents, farm products
\item \code{520} wholesale and retail buyers, except farm products
\item \code{530} purchasing agents, except wholesale, retail, and farm products
\item \code{540} claims adjusters, appraisers, examiners, and investigators
\item \code{565} compliance officers
\item \code{600} cost estimators
\item \code{630} human resources workers
\item \code{640} compensation, benefits, and job analysis specialists
\item \code{650} training and development specialists
\item \code{700} logisticians
\item \code{710} management analysts
\item \code{725} meeting, convention, and event planners
\item \code{726} fundraisers
\item \code{735} market research analysts and marketing specialists
\item \code{740} business operations specialists, all other
\item \code{800} accountants and auditors
\item \code{810} appraisers and assessors of real estate
\item \code{820} budget analysts
\item \code{830} credit analysts
\item \code{840} financial analysts
\item \code{850} personal financial advisors
\item \code{860} insurance underwriters
\item \code{900} financial examiners
\item \code{910} credit counselors and loan officers
\item \code{930} tax examiners and collectors, and revenue agents
\item \code{940} tax preparers
\item \code{950} financial specialists, all other
\item \code{1005} computer and information research scientists
\item \code{1006} computer systems analysts
\item \code{1007} information security analysts
\item \code{1010} computer programmers
\item \code{1020} software developers, applications and systems software
\item \code{1030} web developers
\item \code{1050} computer support specialists
\item \code{1060} database administrators
\item \code{1105} network and computer systems administrators
\item \code{1106} computer network architects
\item \code{1107} computer occupations, all other
\item \code{1200} actuaries
\item \code{1210} mathematicians
\item \code{1220} operations research analysts
\item \code{1230} statisticians
\item \code{1240} miscellaneous mathematical science occupations
\item \code{1300} architects, except naval
\item \code{1310} surveyors, cartographers, and photogrammetrists
\item \code{1320} aerospace engineers
\item \code{1330} agricultural engineers
\item \code{1340} biomedical engineers
\item \code{1350} chemical engineers
\item \code{1360} civil engineers
\item \code{1400} computer hardware engineers
\item \code{1410} electrical and electronics engineers
\item \code{1420} environmental engineers
\item \code{1430} industrial engineers, including health and safety
\item \code{1440} marine engineers and naval architects
\item \code{1450} materials engineers
\item \code{1460} mechanical engineers
\item \code{1500} mining and geological engineers, including mining safety engineers
\item \code{1510} nuclear engineers
\item \code{1520} petroleum engineers
\item \code{1530} engineers, all other
\item \code{1540} drafters
\item \code{1550} engineering technicians, except drafters
\item \code{1560} surveying and mapping technicians
\item \code{1600} agricultural and food scientists
\item \code{1610} biological scientists
\item \code{1640} conservation scientists and foresters
\item \code{1650} medical scientists
\item \code{1660} life scientists, all other
\item \code{1700} astronomers and physicists
\item \code{1710} atmospheric and space scientists
\item \code{1720} chemists and materials scientists
\item \code{1740} environmental scientists and geoscientists
\item \code{1760} physical scientists, all other
\item \code{1800} economists
\item \code{1815} survey researchers
\item \code{1820} psychologists
\item \code{1830} sociologists
\item \code{1840} urban and regional planners
\item \code{1860} miscellaneous social scientists and related workers
\item \code{1900} agricultural and food science technicians
\item \code{1910} biological technicians
\item \code{1920} chemical technicians
\item \code{1930} geological and petroleum technicians
\item \code{1940} nuclear technicians
\item \code{1950} social science research assistants
\item \code{1965} miscellaneous life, physical, and social science technicians
\item \code{2000} counselors
\item \code{2010} social workers
\item \code{2015} probation officers and correctional treatment specialists
\item \code{2016} social and human service assistants
\item \code{2025} miscellaneous community and social service specialists,
\item \code{2040} clergy
\item \code{2050} directors, religious activities and education
\item \code{2060} religious workers, all other
\item \code{2100} lawyers
\item \code{2105} judicial law clerks
\item \code{2110} judges, magistrates, and other judicial workers
\item \code{2145} paralegals and legal assistants
\item \code{2160} miscellaneous legal support workers
\item \code{2200} postsecondary teachers
\item \code{2300} preschool and kindergarten teachers
\item \code{2310} elementary and middle school teachers
\item \code{2320} secondary school teachers
\item \code{2330} special education teachers
\item \code{2340} other teachers and instructors
\item \code{2400} archivists, curators, and museum technicians
\item \code{2430} librarians
\item \code{2440} library technicians
\item \code{2540} teacher assistants
\item \code{2550} other education, training, and library workers
\item \code{2600} artists and related workers
\item \code{2630} designers
\item \code{2700} actors
\item \code{2710} producers and directors
\item \code{2720} athletes, coaches, umpires, and related workers
\item \code{2740} dancers and choreographers
\item \code{2750} musicians, singers, and related workers
\item \code{2760} entertainers and performers, sports and related workers, all other
\item \code{2800} announcers
\item \code{2810} news analysts, reporters and correspondents
\item \code{2825} public relations specialists
\item \code{2830} editors
\item \code{2840} technical writers
\item \code{2850} writers and authors
\item \code{2860} miscellaneous media and communication workers
\item \code{2900} broadcast and sound engineering technicians and radio
\item \code{2910} photographers
\item \code{2920} television, video, and motion picture camera operators and
\item \code{2960} media and communication equipment workers, all other
\item \code{3000} chiropractors
\item \code{3010} dentists
\item \code{3030} dietitians and nutritionists
\item \code{3040} optometrists
\item \code{3050} pharmacists
\item \code{3060} physicians and surgeons
\item \code{3110} physician assistants
\item \code{3120} podiatrists
\item \code{3140} audiologists
\item \code{3150} occupational therapists
\item \code{3160} physical therapists
\item \code{3200} radiation therapists
\item \code{3210} recreational therapists
\item \code{3220} respiratory therapists
\item \code{3230} speech-language pathologists
\item \code{3235} exercise physiologists
\item \code{3245} therapists, all other
\item \code{3250} veterinarians
\item \code{3255} registered nurses
\item \code{3256} nurse anesthetists
\item \code{3257} nurse midwives
\item \code{3258} nurse practitioners
\item \code{3260} health diagnosing and treating practitioners, all other
\item \code{3300} clinical laboratory technologists and technicians
\item \code{3310} dental hygienists
\item \code{3320} diagnostic related technologists and technicians
\item \code{3400} emergency medical technicians and paramedics
\item \code{3420} health practitioner support technologists and technicians
\item \code{3500} licensed practical and licensed vocational nurses
\item \code{3510} medical records and health information technicians
\item \code{3520} opticians, dispensing
\item \code{3535} miscellaneous health technologists and technicians
\item \code{3540} other healthcare practitioners and technical occupations
\item \code{3600} nursing, psychiatric, and home health aides
\item \code{3610} occupational therapy assistants and aides
\item \code{3620} physical therapist assistants and aides
\item \code{3630} massage therapists
\item \code{3640} dental assistants
\item \code{3645} medical assistants
\item \code{3646} medical transcriptionists
\item \code{3647} pharmacy aides
\item \code{3648} veterinary assistants and laboratory animal caretakers
\item \code{3649} phlebotomists
\item \code{3655} miscellaneous healthcare support occupations, including medical equipment preparers
\item \code{3700} first-line supervisors of correctional officers
\item \code{3710} first-line supervisors of police and detectives
\item \code{3720} first-line supervisors of fire fighting and prevention workers
\item \code{3730} first-line supervisors of protective service workers, all other
\item \code{3740} firefighters
\item \code{3750} fire inspectors
\item \code{3800} bailiffs, correctional officers, and jailers
\item \code{3820} detectives and criminal investigators
\item \code{3830} fish and game wardens
\item \code{3840} parking enforcement workers
\item \code{3850} police and sheriff's patrol officers
\item \code{3860} transit and railroad police
\item \code{3900} animal control workers
\item \code{3910} private detectives and investigators
\item \code{3930} security guards and gaming surveillance officers
\item \code{3940} crossing guards
\item \code{3945} transportation security screeners
\item \code{3955} lifeguards and other recreational, and all other protective service workers
\item \code{4000} chefs and head cooks
\item \code{4010} first-line supervisors of food preparation and serving workers
\item \code{4020} cooks
\item \code{4030} food preparation workers
\item \code{4040} bartenders
\item \code{4050} combined food preparation and serving workers, including fast food
\item \code{4060} counter attendants, cafeteria, food concession, and coffee shop
\item \code{4110} waiters and waitresses
\item \code{4120} food servers, nonrestaurant
\item \code{4130} dining room and cafeteria attendants and bartender helpers
\item \code{4140} dishwashers
\item \code{4150} hosts and hostesses, restaurant, lounge, and coffee shop
\item \code{4160} food preparation and serving related workers, all other
\item \code{4200} first-line supervisors of housekeeping and janitorial workers
\item \code{4210} first-line supervisors of landscaping, lawn service, and groundskeeping workers
\item \code{4220} janitors and building cleaners
\item \code{4230} maids and housekeeping cleaners
\item \code{4240} pest control workers
\item \code{4250} grounds maintenance workers
\item \code{4300} first-line supervisors of gaming workers
\item \code{4320} first-line supervisors of personal service workers
\item \code{4340} animal trainers
\item \code{4350} nonfarm animal caretakers
\item \code{4400} gaming services workers
\item \code{4410} motion picture projectionists
\item \code{4420} ushers, lobby attendants, and ticket takers
\item \code{4430} miscellaneous entertainment attendants and related workers
\item \code{4460} embalmers and funeral attendants
\item \code{4465} morticians, undertakers, and funeral directors
\item \code{4500} barbers
\item \code{4510} hairdressers, hairstylists, and cosmetologists
\item \code{4520} miscellaneous personal appearance workers
\item \code{4530} baggage porters, bellhops, and concierges
\item \code{4540} tour and travel guides
\item \code{4600} childcare workers
\item \code{4610} personal care aides
\item \code{4620} recreation and fitness workers
\item \code{4640} residential advisors
\item \code{4650} personal care and service workers, all other
\item \code{4700} first-line supervisors of retail sales workers
\item \code{4710} first-line supervisors of non-retail sales workers
\item \code{4720} cashiers
\item \code{4740} counter and rental clerks
\item \code{4750} parts salespersons
\item \code{4760} retail salespersons
\item \code{4800} advertising sales agents
\item \code{4810} insurance sales agents
\item \code{4820} securities, commodities, and financial services sales agents
\item \code{4830} travel agents
\item \code{4840} sales representatives, services, all other
\item \code{4850} sales representatives, wholesale and manufacturing
\item \code{4900} models, demonstrators, and product promoters
\item \code{4920} real estate brokers and sales agents
\item \code{4930} sales engineers
\item \code{4940} telemarketers
\item \code{4950} door-to-door sales workers, news and street vendors, and related workers
\item \code{4965} sales and related workers, all other
\item \code{5000} first-line supervisors of office and administrative support workers
\item \code{5010} switchboard operators, including answering service
\item \code{5020} telephone operators
\item \code{5030} communications equipment operators, all other
\item \code{5100} bill and account collectors
\item \code{5110} billing and posting clerks
\item \code{5120} bookkeeping, accounting, and auditing clerks
\item \code{5130} gaming cage workers
\item \code{5140} payroll and timekeeping clerks
\item \code{5150} procurement clerks
\item \code{5160} tellers
\item \code{5165} financial clerks, all other
\item \code{5200} brokerage clerks
\item \code{5210} correspondence clerks
\item \code{5220} court, municipal, and license clerks
\item \code{5230} credit authorizers, checkers, and clerks
\item \code{5240} customer service representatives
\item \code{5250} eligibility interviewers, government programs
\item \code{5260} file clerks
\item \code{5300} hotel, motel, and resort desk clerks
\item \code{5310} interviewers, except eligibility and loan
\item \code{5320} library assistants, clerical
\item \code{5330} loan interviewers and clerks
\item \code{5340} new accounts clerks
\item \code{5350} order clerks
\item \code{5360} human resources assistants, except payroll and timekeeping
\item \code{5400} receptionists and information clerks
\item \code{5410} reservation and transportation ticket agents and travel clerks
\item \code{5420} information and record clerks, all other
\item \code{5500} cargo and freight agents
\item \code{5510} couriers and messengers
\item \code{5520} dispatchers
\item \code{5530} meter readers, utilities
\item \code{5540} postal service clerks
\item \code{5550} postal service mail carriers
\item \code{5560} postal service mail sorters, processors, and processing machine operators
\item \code{5600} production, planning, and expediting clerks
\item \code{5610} shipping, receiving, and traffic clerks
\item \code{5620} stock clerks and order fillers
\item \code{5630} weighers, measurers, checkers, and samplers, recordkeeping
\item \code{5700} secretaries and administrative assistants
\item \code{5800} computer operators
\item \code{5810} data entry keyers
\item \code{5820} word processors and typists
\item \code{5830} desktop publishers
\item \code{5840} insurance claims and policy processing clerks
\item \code{5850} mail clerks and mail machine operators, except postal service
\item \code{5860} office clerks, general
\item \code{5900} office machine operators, except computer
\item \code{5910} proofreaders and copy markers
\item \code{5920} statistical assistants
\item \code{5940} office and administrative support workers, all other
\item \code{6005} first-line supervisors of farming, fishing, and forestry workers
\item \code{6010} agricultural inspectors
\item \code{6020} animal breeders
\item \code{6040} graders and sorters, agricultural products
\item \code{6050} miscellaneous agricultural workers
\item \code{6100} fishers and related fishing workers
\item \code{6110} hunters and trappers
\item \code{6120} forest and conservation workers
\item \code{6130} logging workers
\item \code{6200} first-line supervisors of construction trades and extraction workers
\item \code{6210} boilermakers
\item \code{6220} brickmasons, blockmasons, and stonemasons
\item \code{6230} carpenters
\item \code{6240} carpet, floor, and tile installers and finishers
\item \code{6250} cement masons, concrete finishers, and terrazzo workers
\item \code{6260} construction laborers
\item \code{6300} paving, surfacing, and tamping equipment operators
\item \code{6310} pile-driver operators
\item \code{6320} operating engineers and other construction equipment operators
\item \code{6330} drywall installers, ceiling tile installers, and tapers
\item \code{6355} electricians
\item \code{6360} glaziers
\item \code{6400} insulation workers
\item \code{6420} painters, construction and maintenance
\item \code{6430} paperhangers
\item \code{6440} pipelayers, plumbers, pipefitters, and steamfitters
\item \code{6460} plasterers and stucco masons
\item \code{6500} reinforcing iron and rebar workers
\item \code{6515} roofers
\item \code{6520} sheet metal workers
\item \code{6530} structural iron and steel workers
\item \code{6540} solar photovoltaic installers
\item \code{6600} helpers, construction trades
\item \code{6660} construction and building inspectors
\item \code{6700} elevator installers and repairers
\item \code{6710} fence erectors
\item \code{6720} hazardous materials removal workers
\item \code{6730} highway maintenance workers
\item \code{6740} rail-track laying and maintenance equipment operators
\item \code{6750} septic tank servicers and sewer pipe cleaners
\item \code{6765} miscellaneous construction and related workers
\item \code{6800} derrick, rotary drill, and service unit operators, oil, gas, and mining
\item \code{6820} earth drillers, except oil and gas
\item \code{6830} explosives workers, ordnance handling experts, and blasters
\item \code{6840} mining machine operators
\item \code{6910} roof bolters, mining
\item \code{6920} roustabouts, oil and gas
\item \code{6930} helpers?xtraction workers
\item \code{6940} other extraction workers
\item \code{7000} first-line supervisors of mechanics, installers, and repairers
\item \code{7010} computer, automated teller, and office machine repairers
\item \code{7020} radio and telecommunications equipment installers and repairers
\item \code{7030} avionics technicians
\item \code{7040} electric motor, power tool, and related repairers
\item \code{7050} electrical and electronics installers and repairers, transportation
\item \code{7100} electrical and electronics repairers, industrial and utility
\item \code{7110} electronic equipment installers and repairers, motor vehicles
\item \code{7120} electronic home entertainment equipment installers and repairers
\item \code{7130} security and fire alarm systems installers
\item \code{7140} aircraft mechanics and service technicians
\item \code{7150} automotive body and related repairers
\item \code{7160} automotive glass installers and repairers
\item \code{7200} automotive service technicians and mechanics
\item \code{7210} bus and truck mechanics and diesel engine specialists
\item \code{7220} heavy vehicle and mobile equipment service technicians and mechanics
\item \code{7240} small engine mechanics
\item \code{7260} miscellaneous vehicle and mobile equipment mechanics, installers, and repairers
\item \code{7300} control and valve installers and repairers
\item \code{7315} heating, air conditioning, and refrigeration mechanics and installers
\item \code{7320} home appliance repairers
\item \code{7330} industrial and refractory machinery mechanics
\item \code{7340} maintenance and repair workers, general
\item \code{7350} maintenance workers, machinery
\item \code{7360} millwrights
\item \code{7410} electrical power-line installers and repairers
\item \code{7420} telecommunications line installers and repairers
\item \code{7430} precision instrument and equipment repairers
\item \code{7440} wind turbine service technicians
\item \code{7510} coin, vending, and amusement machine servicers and repairers
\item \code{7520} commercial divers
\item \code{7540} locksmiths and safe repairers
\item \code{7550} manufactured building and mobile home installers
\item \code{7560} riggers
\item \code{7600} signal and track switch repairers
\item \code{7610} helpers?nstallation, maintenance, and repair workers
\item \code{7630} other installation, maintenance, and repair workers
\item \code{7700} first-line supervisors of production and operating workers
\item \code{7710} aircraft structure, surfaces, rigging, and systems assemblers
\item \code{7720} electrical, electronics, and electromechanical assemblers
\item \code{7730} engine and other machine assemblers
\item \code{7740} structural metal fabricators and fitters
\item \code{7750} miscellaneous assemblers and fabricators
\item \code{7800} bakers
\item \code{7810} butchers and other meat, poultry, and fish processing workers
\item \code{7830} food and tobacco roasting, baking, and drying machine operators and tenders
\item \code{7840} food batchmakers
\item \code{7850} food cooking machine operators and tenders
\item \code{7855} food processing workers, all other
\item \code{7900} computer control programmers and operators
\item \code{7920} extruding and drawing machine setters, operators, and tenders, metal and plastic
\item \code{7930} forging machine setters, operators, and tenders, metal and plastic
\item \code{7940} rolling machine setters, operators, and tenders, metal and plastic
\item \code{7950} cutting, punching, and press machine setters, operators, and tenders, metal and plastic
\item \code{7960} drilling and boring machine tool setters, operators, and tenders, metal and plastic
\item \code{8000} grinding, lapping, polishing, and buffing machine tool setters, operators, and tenders, metal and plastic
\item \code{8010} lathe and turning machine tool setters, operators, and tenders, metal and plastic
\item \code{8020} milling and planing machine setters, operators, and tenders, metal and plastic
\item \code{8030} machinists
\item \code{8040} metal furnace operators, tenders, pourers, and casters
\item \code{8060} model makers and patternmakers, metal and plastic
\item \code{8100} molders and molding machine setters, operators, and tenders, metal and plastic
\item \code{8120} multiple machine tool setters, operators, and tenders, metal and plastic
\item \code{8130} tool and die makers
\item \code{8140} welding, soldering, and brazing workers
\item \code{8150} heat treating equipment setters, operators, and tenders, metal and plastic
\item \code{8160} layout workers, metal and plastic
\item \code{8200} plating and coating machine setters, operators, and tenders, metal and plastic
\item \code{8210} tool grinders, filers, and sharpeners
\item \code{8220} metal workers and plastic workers, all other
\item \code{8250} prepress technicians and workers
\item \code{8255} printing press operators
\item \code{8256} print binding and finishing workers
\item \code{8300} laundry and dry-cleaning workers
\item \code{8310} pressers, textile, garment, and related materials
\item \code{8320} sewing machine operators
\item \code{8330} shoe and leather workers and repairers
\item \code{8340} shoe machine operators and tenders
\item \code{8350} tailors, dressmakers, and sewers
\item \code{8360} textile bleaching and dyeing machine operators and tenders
\item \code{8400} textile cutting machine setters, operators, and tenders
\item \code{8410} textile knitting and weaving machine setters, operators, and tenders
\item \code{8420} textile winding, twisting, and drawing out machine setters, operators, and tenders
\item \code{8430} extruding and forming machine setters, operators, and tenders, synthetic and glass fibers
\item \code{8440} fabric and apparel patternmakers
\item \code{8450} upholsterers
\item \code{8460} textile, apparel, and furnishings workers, all other
\item \code{8500} cabinetmakers and bench carpenters
\item \code{8510} furniture finishers
\item \code{8520} model makers and patternmakers, wood
\item \code{8530} sawing machine setters, operators, and tenders, wood
\item \code{8540} woodworking machine setters, operators, and tenders, except sawing
\item \code{8550} woodworkers, all other
\item \code{8600} power plant operators, distributors, and dispatchers
\item \code{8610} stationary engineers and boiler operators
\item \code{8620} water and wastewater treatment plant and system operators
\item \code{8630} miscellaneous plant and system operators
\item \code{8640} chemical processing machine setters, operators, and tenders
\item \code{8650} crushing, grinding, polishing, mixing, and blending workers
\item \code{8710} cutting workers
\item \code{8720} extruding, forming, pressing, and compacting machine setters,
\item \code{8730} operators, and tenders furnace, kiln, oven, drier, and kettle operators and tenders
\item \code{8740} inspectors, testers, sorters, samplers, and weighers
\item \code{8750} jewelers and precious stone and metal workers
\item \code{8760} medical, dental, and ophthalmic laboratory technicians
\item \code{8800} packaging and filling machine operators and tenders
\item \code{8810} painting workers
\item \code{8830} photographic process workers and processing machine operators
\item \code{8840} semiconductor processors
\item \code{8850} adhesive bonding machine operators and tenders
\item \code{8860} cleaning, washing, and metal pickling equipment operators and tenders
\item \code{8900} cooling and freezing equipment operators and tenders
\item \code{8910} etchers and engravers
\item \code{8920} molders, shapers, and casters, except metal and plastic
\item \code{8930} paper goods machine setters, operators, and tenders
\item \code{8940} tire builders
\item \code{8950} helpers?roduction workers
\item \code{8965} production workers, all other
\item \code{9000} supervisors of transportation and material moving workers
\item \code{9030} aircraft pilots and flight engineers
\item \code{9040} air traffic controllers and airfield operations specialists
\item \code{9050} flight attendants
\item \code{9110} ambulance drivers and attendants, except emergency medical technicians
\item \code{9120} bus drivers
\item \code{9130} driver/sales workers and truck drivers
\item \code{9140} taxi drivers and chauffeurs
\item \code{9150} motor vehicle operators, all other
\item \code{9200} locomotive engineers and operators
\item \code{9230} railroad brake, signal, and switch operators
\item \code{9240} railroad conductors and yardmasters
\item \code{9260} subway, streetcar, and other rail transportation workers
\item \code{9300} sailors and marine oilers
\item \code{9310} ship and boat captains and operators
\item \code{9330} ship engineers
\item \code{9340} bridge and lock tenders
\item \code{9350} parking lot attendants
\item \code{9360} automotive and watercraft service attendants
\item \code{9410} transportation inspectors
\item \code{9415} transportation attendants, except flight attendants
\item \code{9420} other transportation workers
\item \code{9500} conveyor operators and tenders
\item \code{9510} crane and tower operators
\item \code{9520} dredge, excavating, and loading machine operators
\item \code{9560} hoist and winch operators
\item \code{9600} industrial truck and tractor operators
\item \code{9610} cleaners of vehicles and equipment
\item \code{9620} laborers and freight, stock, and material movers, hand
\item \code{9630} machine feeders and offbearers
\item \code{9640} packers and packagers, hand
\item \code{9650} pumping station operators
\item \code{9720} refuse and recyclable material collectors
\item \code{9730} mine shuttle car operators
\item \code{9740} tank car, truck, and ship loaders
\item \code{9750} material moving workers, all other
\item \code{9800} military officer special and tactical operations leaders
\item \code{9810} first-line enlisted military supervisors
\item \code{9820} military enlisted tactical operations and air/weapons specialists and crew members
\item \code{9830} military, rank not specified
\item \code{9997} uncodeable
\item \code{9998} dk
\item \code{9999} na
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
