% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_368.R
\name{scitext}
\alias{scitext}
\title{What it means to r to study scientifically}
\source{
General Social Survey https://gss.norc.org
}
\description{
scitext
}
\details{
Question 1040. In your own words, could you tell me what it means to study something scientifically?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab classification \tab do experiments, control group \tab don't know \tab formulation of theories, test hypotheses \tab measurement \tab no answer \tab redundancies/incorrect \tab rigorous, systematic comparison \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2995 \tab 400 \tab 165 \tab 127 \tab 206 \tab 292 \tab 14 \tab 218 \tab 93 \tab 4510 \cr
   2008 \tab 849 \tab 327 \tab 112 \tab 93 \tab 125 \tab 197 \tab 23 \tab 215 \tab 82 \tab 2023 \cr
   2010 \tab 1480 \tab 176 \tab 43 \tab 46 \tab 57 \tab 72 \tab 13 \tab 110 \tab 47 \tab 2044 \cr
   2012 \tab 1222 \tab 262 \tab 58 \tab 65 \tab 71 \tab 76 \tab 19 \tab 164 \tab 37 \tab 1974 \cr
   2014 \tab 1572 \tab 358 \tab 111 \tab 54 \tab 105 \tab 46 \tab 26 \tab 177 \tab 89 \tab 2538 \cr
   2016 \tab 1766 \tab 337 \tab 114 \tab 80 \tab 119 \tab 116 \tab 25 \tab 232 \tab 78 \tab 2867 \cr
   2018 \tab 1418 \tab 262 \tab 85 \tab 71 \tab 117 \tab 158 \tab 10 \tab 126 \tab 101 \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 65388 \tab 2122 \tab 688 \tab 536 \tab 800 \tab 957 \tab 130 \tab 1242 \tab 527 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} formulation of theories, test hypotheses
\item \code{2} do experiments, control group
\item \code{3} rigorous, systematic comparison
\item \code{4} measurement
\item \code{5} classification
\item \code{6} redundancies/incorrect
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
