% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_596.R
\name{sei}
\alias{sei}
\title{sei}
\source{
General Social Survey https://gss.norc.org
}
\description{
Respondent socioeconomic index (1980)
}
\details{
Question 1646. Respondent socioeconomic index.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 17.1 \tab 17.6 \tab 18.5 \tab 19.4 \tab 20.1 \tab 21.2 \tab 22.5 \tab 22.6 \tab 22.7 \tab 22.9 \tab 23.3 \tab 24.2 \tab 24.6 \tab 25.2 \tab 25.3 \tab 25.4 \tab 25.5 \tab 25.8 \tab 25.9 \tab 26.4 \tab 26.5 \tab 26.6 \tab 26.7 \tab 26.9 \tab 27.2 \tab 27.5 \tab 28 \tab 28.2 \tab 28.4 \tab 28.6 \tab 29 \tab 29.1 \tab 29.2 \tab 29.3 \tab 29.4 \tab 29.5 \tab 29.6 \tab 29.7 \tab 29.8 \tab 30 \tab 30.1 \tab 30.4 \tab 30.5 \tab 30.7 \tab 30.8 \tab 31.2 \tab 31.3 \tab 31.4 \tab 31.5 \tab 31.7 \tab 31.9 \tab 32 \tab 32.1 \tab 32.3 \tab 32.4 \tab 32.7 \tab 32.8 \tab 32.9 \tab 33.1 \tab 33.2 \tab 33.3 \tab 33.5 \tab 33.6 \tab 33.8 \tab 33.9 \tab 34.2 \tab 34.3 \tab 34.4 \tab 34.6 \tab 34.8 \tab 35.1 \tab 35.2 \tab 35.4 \tab 36.1 \tab 36.4 \tab 36.5 \tab 36.9 \tab 37 \tab 37.3 \tab 37.4 \tab 37.6 \tab 37.7 \tab 37.8 \tab 37.9 \tab 38 \tab 38.1 \tab 38.2 \tab 38.4 \tab 38.5 \tab 38.8 \tab 38.9 \tab 39 \tab 39.4 \tab 39.5 \tab 39.6 \tab 40 \tab 40.1 \tab 41.1 \tab 41.2 \tab 41.3 \tab 41.6 \tab 42 \tab 42.3 \tab 42.4 \tab 42.5 \tab 42.7 \tab 43.1 \tab 43.4 \tab 43.5 \tab 43.6 \tab 43.7 \tab 44.4 \tab 44.7 \tab 44.8 \tab 45.2 \tab 45.3 \tab 45.4 \tab 45.7 \tab 45.8 \tab 45.9 \tab 46.1 \tab 46.3 \tab 46.7 \tab 46.8 \tab 47 \tab 48.5 \tab 48.6 \tab 48.8 \tab 48.9 \tab 49.4 \tab 49.6 \tab 49.7 \tab 49.9 \tab 50.2 \tab 50.5 \tab 50.7 \tab 51 \tab 51.8 \tab 51.9 \tab 52.3 \tab 52.4 \tab 52.5 \tab 53.1 \tab 53.3 \tab 53.4 \tab 53.7 \tab 53.9 \tab 54.2 \tab 54.7 \tab 55.1 \tab 55.3 \tab 56.5 \tab 56.6 \tab 57.1 \tab 58.3 \tab 58.8 \tab 59.1 \tab 60.3 \tab 60.5 \tab 60.6 \tab 61.1 \tab 61.4 \tab 61.9 \tab 62 \tab 62.2 \tab 62.5 \tab 63.2 \tab 63.5 \tab 64 \tab 64.1 \tab 64.2 \tab 64.4 \tab 64.8 \tab 65.2 \tab 65.8 \tab 66.1 \tab 67 \tab 68.1 \tab 68.8 \tab 69.2 \tab 69.7 \tab 71.6 \tab 72.5 \tab 73.2 \tab 73.3 \tab 73.6 \tab 73.8 \tab 73.9 \tab 74.4 \tab 74.9 \tab 75.2 \tab 75.9 \tab 76.3 \tab 76.4 \tab 78 \tab 78.5 \tab 78.7 \tab 78.8 \tab 80 \tab 80.3 \tab 80.9 \tab 81 \tab 81.3 \tab 81.9 \tab 82.7 \tab 83.7 \tab 84 \tab 85 \tab 85.1 \tab 85.5 \tab 85.7 \tab 85.8 \tab 86 \tab 86.5 \tab 86.6 \tab 86.7 \tab 86.8 \tab 87 \tab 87.4 \tab 87.9 \tab 88.2 \tab 88.3 \tab 88.8 \tab 92.3 \tab 92.8 \tab 93.3 \tab 96 \tab 97.2 \tab 19.5 \tab 21.1 \tab 21.6 \tab 24.8 \tab 27.7 \tab 27.9 \tab 28.8 \tab 28.9 \tab 30.2 \tab 33 \tab 33.4 \tab 36.2 \tab 36.7 \tab 37.1 \tab 37.5 \tab 38.7 \tab 39.7 \tab 39.8 \tab 40.4 \tab 40.9 \tab 42.2 \tab 43.2 \tab 44.3 \tab 44.5 \tab 47.8 \tab 49.3 \tab 53.2 \tab 53.5 \tab 58.1 \tab 59 \tab 59.3 \tab 59.9 \tab 62.1 \tab 63.9 \tab 66.3 \tab 70.5 \tab 71.5 \tab 74.1 \tab 74.6 \tab 79.3 \tab 79.9 \tab 83.9 \tab 84.2 \tab 84.7 \tab 89.2 \tab 89.3 \tab 93.1 \tab 26.3 \tab 31.1 \tab 36 \tab 37.2 \tab 39.9 \tab 44.2 \tab 50.3 \tab 54.5 \tab 54.9 \tab 71.2 \tab 73.7 \tab 74 \tab 76.5 \tab 79.8 \tab 83.2 \tab 86.9 \tab 87.2 \tab 87.3 \tab 87.5 \tab 20.7 \tab 27.3 \tab 34 \tab 35.7 \tab 40.2 \tab 42.9 \tab 45 \tab 46 \tab 48.4 \tab 84.5 \tab 90.3 \tab 91.2 \tab 21.3 \tab 47.6 \tab 60.4 \tab 90.1 \tab 27.4 \tab 34.7 \tab 40.5 \tab 42.1 \tab 43.3 \tab 56.4 \tab 72 \tab 75.3 \tab 78.9 \tab 88.6 \tab 90 \tab 70 \tab 77.8 \tab 88.9 \tab 90.8 \tab 92.9 \tab 30.6 \tab 38.3 \tab 46.2 \tab 91.3 \tab 80.1 \tab 93.7 \tab 91.7 \tab 71.3 \tab 29.9 \tab 80.2 \tab 45.1 \tab 47.9 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 94 \tab 23 \tab 3 \tab 1 \tab 2 \tab 2 \tab 10 \tab 14 \tab 5 \tab 4 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 12 \tab 6 \tab 2 \tab 24 \tab 2 \tab 4 \tab 19 \tab 2 \tab 2 \tab 24 \tab 15 \tab 6 \tab 5 \tab 21 \tab 27 \tab 10 \tab 2 \tab 3 \tab 1 \tab 11 \tab 1 \tab 2 \tab 1 \tab 1 \tab 9 \tab 16 \tab 2 \tab 5 \tab 12 \tab 13 \tab 1 \tab 1 \tab 9 \tab 1 \tab 18 \tab 10 \tab 1 \tab 1 \tab 1 \tab 45 \tab 1 \tab 4 \tab 1 \tab 8 \tab 1 \tab 24 \tab 18 \tab 4 \tab 2 \tab 3 \tab 9 \tab 10 \tab 7 \tab 9 \tab 5 \tab 4 \tab 33 \tab 4 \tab 2 \tab 24 \tab 6 \tab 7 \tab 21 \tab 5 \tab 3 \tab 26 \tab 2 \tab 7 \tab 72 \tab 5 \tab 2 \tab 30 \tab 5 \tab 5 \tab 3 \tab 1 \tab 7 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 8 \tab 2 \tab 7 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 5 \tab 3 \tab 1 \tab 1 \tab 2 \tab 5 \tab 5 \tab 11 \tab 2 \tab 5 \tab 5 \tab 1 \tab 10 \tab 1 \tab 3 \tab 3 \tab 7 \tab 23 \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 5 \tab 6 \tab 3 \tab 1 \tab 1 \tab 6 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 8 \tab 1 \tab 1 \tab 5 \tab 7 \tab 2 \tab 10 \tab 13 \tab 85 \tab 1 \tab 3 \tab 8 \tab 2 \tab 7 \tab 1 \tab 5 \tab 6 \tab 10 \tab 5 \tab 1 \tab 4 \tab 4 \tab 2 \tab 2 \tab 22 \tab 5 \tab 9 \tab 2 \tab 5 \tab 3 \tab 1 \tab 3 \tab 1 \tab 6 \tab 13 \tab 4 \tab 27 \tab 1 \tab 1 \tab 1 \tab 8 \tab 4 \tab 2 \tab 1 \tab 1 \tab 3 \tab 8 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 2 \tab 2 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 113 \tab 11 \tab 4 \tab 2 \tab 1 \tab 1 \tab 11 \tab 11 \tab 4 \tab 3 \tab 8 \tab - \tab 1 \tab 1 \tab 4 \tab 5 \tab 8 \tab 3 \tab 5 \tab 1 \tab 10 \tab 12 \tab 1 \tab 10 \tab 1 \tab 1 \tab 10 \tab - \tab 3 \tab 24 \tab 12 \tab 3 \tab 5 \tab 17 \tab 26 \tab 10 \tab 1 \tab 3 \tab - \tab 7 \tab 3 \tab 3 \tab 2 \tab 1 \tab 4 \tab 6 \tab 2 \tab 3 \tab 16 \tab 16 \tab - \tab 2 \tab 7 \tab 1 \tab 21 \tab 17 \tab 2 \tab - \tab - \tab 42 \tab 2 \tab 3 \tab - \tab 10 \tab 1 \tab 8 \tab 11 \tab 5 \tab 4 \tab - \tab 4 \tab 11 \tab 11 \tab 9 \tab 2 \tab 3 \tab 23 \tab 1 \tab 1 \tab 31 \tab 6 \tab 3 \tab 27 \tab 8 \tab 6 \tab 16 \tab 8 \tab 4 \tab 92 \tab 6 \tab 5 \tab 19 \tab 3 \tab 7 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 15 \tab 1 \tab 7 \tab - \tab 1 \tab 2 \tab 6 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 8 \tab 1 \tab - \tab 8 \tab 5 \tab 1 \tab 1 \tab - \tab 7 \tab - \tab - \tab - \tab 9 \tab 39 \tab 7 \tab - \tab 9 \tab 1 \tab 2 \tab 6 \tab 1 \tab 8 \tab 1 \tab 6 \tab 2 \tab 1 \tab 3 \tab 3 \tab 3 \tab 2 \tab 5 \tab 4 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab 5 \tab - \tab 1 \tab 13 \tab 2 \tab 2 \tab 13 \tab 15 \tab 80 \tab - \tab 15 \tab 5 \tab 5 \tab 6 \tab - \tab 8 \tab 5 \tab 5 \tab 4 \tab 2 \tab 7 \tab 7 \tab 4 \tab 3 \tab 27 \tab 6 \tab 5 \tab 2 \tab 2 \tab 7 \tab 3 \tab 4 \tab 1 \tab 3 \tab 12 \tab - \tab 23 \tab - \tab - \tab - \tab 13 \tab 8 \tab 1 \tab 1 \tab 2 \tab - \tab 5 \tab 1 \tab - \tab 8 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 5 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 11 \tab 1 \tab 2 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 2 \tab 3 \tab 1 \tab 1 \tab 3 \tab 4 \tab 1 \tab 2 \tab 1 \tab 5 \tab 4 \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 6 \tab 4 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 78 \tab 15 \tab 3 \tab 2 \tab 2 \tab 1 \tab 5 \tab 10 \tab - \tab 11 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab 2 \tab 4 \tab - \tab - \tab 13 \tab 11 \tab 2 \tab 15 \tab - \tab - \tab 15 \tab 1 \tab 2 \tab 14 \tab 24 \tab 5 \tab 7 \tab 18 \tab 25 \tab 15 \tab 1 \tab 5 \tab - \tab 6 \tab 1 \tab 5 \tab 1 \tab - \tab 3 \tab 11 \tab 2 \tab 4 \tab 8 \tab 26 \tab - \tab 1 \tab 15 \tab - \tab 9 \tab 15 \tab 4 \tab - \tab 1 \tab 36 \tab 1 \tab 4 \tab - \tab 9 \tab 1 \tab 13 \tab 15 \tab 7 \tab - \tab 6 \tab 3 \tab 7 \tab 2 \tab 4 \tab 2 \tab - \tab 25 \tab 6 \tab 2 \tab 22 \tab 4 \tab 6 \tab 22 \tab 7 \tab 7 \tab 21 \tab 9 \tab 7 \tab 70 \tab 3 \tab 2 \tab 25 \tab 5 \tab 3 \tab 1 \tab 1 \tab 6 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 7 \tab 1 \tab 5 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 4 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 3 \tab 11 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 5 \tab 26 \tab - \tab 1 \tab 13 \tab 2 \tab 2 \tab 2 \tab 1 \tab 8 \tab - \tab 3 \tab 2 \tab 7 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 7 \tab 7 \tab - \tab - \tab 8 \tab 3 \tab 5 \tab 10 \tab 14 \tab 58 \tab - \tab 12 \tab 14 \tab 1 \tab 5 \tab - \tab 8 \tab 10 \tab 4 \tab 1 \tab 1 \tab 13 \tab 2 \tab 5 \tab 1 \tab 24 \tab 3 \tab 3 \tab 5 \tab 2 \tab 4 \tab - \tab 4 \tab 2 \tab 2 \tab 18 \tab 2 \tab 34 \tab 1 \tab - \tab 2 \tab 14 \tab 5 \tab 1 \tab - \tab 2 \tab 5 \tab 5 \tab - \tab 1 \tab 6 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 8 \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 6 \tab - \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 111 \tab 23 \tab - \tab 7 \tab 3 \tab 1 \tab 5 \tab 5 \tab - \tab 6 \tab 4 \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab 2 \tab 4 \tab 4 \tab - \tab 12 \tab 7 \tab 1 \tab 9 \tab - \tab 2 \tab 24 \tab 1 \tab 2 \tab 21 \tab 18 \tab 4 \tab 4 \tab 7 \tab 35 \tab 15 \tab 7 \tab - \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 2 \tab 4 \tab 4 \tab - \tab - \tab 16 \tab 12 \tab - \tab 2 \tab 15 \tab 4 \tab 15 \tab 24 \tab 7 \tab - \tab 1 \tab 51 \tab 3 \tab 3 \tab - \tab 5 \tab 2 \tab 16 \tab 16 \tab 5 \tab 2 \tab 5 \tab 5 \tab 9 \tab 1 \tab 6 \tab 3 \tab 5 \tab 28 \tab 6 \tab 2 \tab 21 \tab 8 \tab 8 \tab 14 \tab 3 \tab 2 \tab 24 \tab 11 \tab 6 \tab 71 \tab 3 \tab 1 \tab 18 \tab 6 \tab 10 \tab 3 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 14 \tab 1 \tab 5 \tab 2 \tab 6 \tab 3 \tab 5 \tab 2 \tab 5 \tab 3 \tab 1 \tab 1 \tab 9 \tab 1 \tab 3 \tab 13 \tab - \tab 2 \tab 4 \tab - \tab 8 \tab - \tab - \tab - \tab 7 \tab 38 \tab 4 \tab - \tab 13 \tab 6 \tab 2 \tab 1 \tab 4 \tab 8 \tab - \tab 3 \tab 4 \tab 1 \tab 4 \tab 2 \tab - \tab 2 \tab 8 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 4 \tab 4 \tab - \tab - \tab 7 \tab 3 \tab 2 \tab 10 \tab 12 \tab 65 \tab - \tab 14 \tab 6 \tab 2 \tab 5 \tab - \tab 11 \tab 6 \tab 5 \tab 2 \tab 5 \tab 9 \tab 12 \tab - \tab 3 \tab 21 \tab 6 \tab 7 \tab 3 \tab - \tab 4 \tab 2 \tab 5 \tab - \tab 3 \tab 19 \tab 2 \tab 38 \tab 2 \tab 1 \tab 1 \tab 16 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 4 \tab 1 \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 7 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 9 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 6 \tab 3 \tab - \tab 1 \tab 6 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 89 \tab 23 \tab 4 \tab - \tab 1 \tab 1 \tab 10 \tab 12 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab 24 \tab 8 \tab 1 \tab 17 \tab 1 \tab 2 \tab 20 \tab 2 \tab 6 \tab 20 \tab 13 \tab 2 \tab 6 \tab 31 \tab 31 \tab 12 \tab 1 \tab 2 \tab - \tab 7 \tab 3 \tab 2 \tab 1 \tab - \tab 6 \tab 5 \tab 2 \tab 5 \tab 19 \tab 24 \tab - \tab 3 \tab 20 \tab 1 \tab 16 \tab 27 \tab 4 \tab - \tab - \tab 52 \tab - \tab 3 \tab - \tab 7 \tab - \tab 4 \tab 12 \tab 6 \tab 1 \tab 12 \tab 2 \tab 3 \tab - \tab 10 \tab 6 \tab 1 \tab 29 \tab 4 \tab 1 \tab 32 \tab 6 \tab 4 \tab 19 \tab 4 \tab 4 \tab 26 \tab 8 \tab 1 \tab 66 \tab 4 \tab 1 \tab 15 \tab 5 \tab 9 \tab 2 \tab 6 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 4 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 9 \tab 1 \tab 12 \tab 1 \tab 1 \tab 4 \tab 4 \tab 1 \tab 3 \tab 4 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 11 \tab 4 \tab 1 \tab 10 \tab 1 \tab 4 \tab - \tab 1 \tab 3 \tab 5 \tab 55 \tab 7 \tab 4 \tab 12 \tab 4 \tab 3 \tab 3 \tab 2 \tab 12 \tab - \tab 10 \tab 3 \tab 3 \tab 5 \tab 3 \tab 1 \tab 2 \tab 4 \tab 3 \tab 5 \tab - \tab 1 \tab 1 \tab 6 \tab 6 \tab - \tab - \tab 1 \tab 3 \tab 3 \tab 11 \tab 17 \tab 81 \tab 1 \tab 20 \tab 11 \tab 9 \tab 6 \tab 2 \tab 13 \tab 5 \tab 14 \tab - \tab 1 \tab 11 \tab 5 \tab 3 \tab 2 \tab 23 \tab 1 \tab 6 \tab 2 \tab 3 \tab 4 \tab 1 \tab 2 \tab 5 \tab 4 \tab 18 \tab 1 \tab 30 \tab - \tab - \tab 2 \tab 12 \tab 6 \tab 4 \tab 1 \tab 2 \tab 3 \tab 7 \tab - \tab - \tab 9 \tab - \tab - \tab - \tab 3 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 1 \tab 16 \tab 2 \tab 3 \tab 2 \tab 7 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab 7 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 163 \tab 22 \tab 3 \tab 2 \tab 2 \tab 1 \tab 20 \tab 25 \tab 2 \tab 17 \tab 6 \tab 1 \tab 6 \tab 1 \tab 8 \tab 3 \tab 1 \tab - \tab 4 \tab 2 \tab 27 \tab 8 \tab 3 \tab 27 \tab 2 \tab 4 \tab 30 \tab - \tab 3 \tab 44 \tab 26 \tab 16 \tab 5 \tab 36 \tab 76 \tab 22 \tab 12 \tab 4 \tab 3 \tab 14 \tab 2 \tab 7 \tab 2 \tab - \tab 14 \tab 9 \tab - \tab 8 \tab 33 \tab 56 \tab - \tab 4 \tab 19 \tab 3 \tab 35 \tab 51 \tab 10 \tab - \tab - \tab 117 \tab 7 \tab 14 \tab - \tab 22 \tab - \tab 21 \tab 20 \tab 18 \tab 2 \tab 9 \tab 4 \tab 9 \tab 1 \tab 22 \tab 5 \tab 1 \tab 49 \tab 12 \tab 4 \tab 38 \tab 10 \tab 13 \tab 27 \tab 2 \tab 7 \tab 47 \tab 20 \tab 6 \tab 132 \tab 5 \tab 1 \tab 26 \tab 13 \tab 20 \tab 5 \tab 2 \tab 8 \tab 2 \tab 1 \tab - \tab 1 \tab 5 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 23 \tab 2 \tab 16 \tab - \tab 3 \tab 2 \tab 8 \tab 1 \tab 6 \tab 4 \tab 1 \tab 3 \tab 5 \tab 3 \tab 6 \tab 20 \tab 3 \tab 3 \tab 6 \tab 1 \tab 7 \tab - \tab 4 \tab 4 \tab 30 \tab 84 \tab 4 \tab 1 \tab 25 \tab 9 \tab 4 \tab 6 \tab 8 \tab 18 \tab 3 \tab 7 \tab 10 \tab 13 \tab 8 \tab - \tab 2 \tab 8 \tab 6 \tab 22 \tab 3 \tab 1 \tab - \tab - \tab 13 \tab 10 \tab 2 \tab 4 \tab 8 \tab 5 \tab 11 \tab 21 \tab 36 \tab 112 \tab - \tab 34 \tab 18 \tab 7 \tab 13 \tab 2 \tab 12 \tab 8 \tab 17 \tab 3 \tab 5 \tab 29 \tab 5 \tab 3 \tab 5 \tab 44 \tab 17 \tab 5 \tab 4 \tab 6 \tab 10 \tab 3 \tab 8 \tab 6 \tab 17 \tab 36 \tab 7 \tab 68 \tab 2 \tab - \tab 4 \tab 24 \tab 10 \tab 9 \tab 2 \tab 3 \tab 12 \tab 21 \tab 2 \tab 1 \tab 19 \tab - \tab 1 \tab - \tab 6 \tab 5 \tab 11 \tab 2 \tab 5 \tab 2 \tab - \tab 5 \tab - \tab 2 \tab 3 \tab 14 \tab 3 \tab 1 \tab 1 \tab 9 \tab 2 \tab 1 \tab - \tab 2 \tab 7 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab 6 \tab 3 \tab - \tab 3 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 9 \tab 4 \tab 5 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 5 \tab 2 \tab 4 \tab 9 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 6 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 3 \tab 5 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 123 \tab 24 \tab 3 \tab 5 \tab 1 \tab 4 \tab 19 \tab 17 \tab 7 \tab 11 \tab 5 \tab 2 \tab 4 \tab - \tab 6 \tab 2 \tab 1 \tab 2 \tab 6 \tab - \tab 33 \tab 11 \tab 3 \tab 37 \tab 1 \tab 4 \tab 31 \tab - \tab 3 \tab 54 \tab 33 \tab 18 \tab 5 \tab 51 \tab 55 \tab 18 \tab 10 \tab 3 \tab 1 \tab 15 \tab 1 \tab 8 \tab - \tab - \tab 7 \tab 11 \tab 4 \tab 8 \tab 25 \tab 39 \tab 2 \tab 2 \tab 14 \tab 5 \tab 37 \tab 33 \tab 8 \tab 3 \tab - \tab 81 \tab 4 \tab 8 \tab - \tab 18 \tab - \tab 18 \tab 24 \tab 13 \tab 6 \tab 6 \tab 1 \tab 20 \tab 3 \tab 20 \tab 3 \tab 3 \tab 69 \tab 11 \tab 1 \tab 35 \tab 13 \tab 10 \tab 33 \tab 3 \tab 10 \tab 43 \tab 19 \tab 17 \tab 97 \tab 2 \tab - \tab 56 \tab 17 \tab 15 \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab - \tab - \tab 4 \tab - \tab 4 \tab 2 \tab 2 \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab 20 \tab - \tab 10 \tab 1 \tab 1 \tab 7 \tab 4 \tab - \tab 8 \tab 4 \tab 1 \tab 1 \tab 5 \tab 4 \tab 3 \tab 31 \tab 9 \tab 2 \tab 11 \tab 1 \tab 12 \tab - \tab 4 \tab - \tab 19 \tab 80 \tab 8 \tab 1 \tab 10 \tab 5 \tab 5 \tab 9 \tab 3 \tab 17 \tab 5 \tab 20 \tab 10 \tab 6 \tab 5 \tab 4 \tab - \tab 7 \tab 9 \tab 10 \tab 6 \tab 1 \tab 4 \tab - \tab 10 \tab 14 \tab 4 \tab 2 \tab 13 \tab 4 \tab 5 \tab 18 \tab 30 \tab 182 \tab - \tab 10 \tab 11 \tab 8 \tab 7 \tab 2 \tab 11 \tab 13 \tab 15 \tab 13 \tab 6 \tab 22 \tab 23 \tab 6 \tab 4 \tab 47 \tab 12 \tab 13 \tab 7 \tab 10 \tab 19 \tab 5 \tab 3 \tab 1 \tab 12 \tab 26 \tab 4 \tab 62 \tab 2 \tab - \tab 3 \tab 31 \tab 10 \tab 5 \tab - \tab 2 \tab 14 \tab 21 \tab 5 \tab - \tab 10 \tab 1 \tab - \tab 1 \tab 3 \tab 10 \tab 13 \tab - \tab 3 \tab 5 \tab 1 \tab 9 \tab 1 \tab 2 \tab 4 \tab 10 \tab - \tab 2 \tab 4 \tab 15 \tab - \tab 4 \tab - \tab 2 \tab 13 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 4 \tab 2 \tab 4 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab 4 \tab 10 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 5 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 160 \tab 9 \tab 1 \tab 2 \tab 1 \tab 2 \tab 14 \tab 12 \tab 5 \tab 14 \tab 9 \tab 2 \tab 1 \tab - \tab 3 \tab 8 \tab - \tab 8 \tab 6 \tab - \tab 22 \tab 6 \tab 4 \tab 43 \tab 1 \tab 2 \tab 36 \tab 1 \tab 1 \tab 43 \tab 30 \tab 11 \tab 5 \tab 34 \tab 59 \tab 29 \tab 5 \tab - \tab 1 \tab 10 \tab 1 \tab 7 \tab 1 \tab - \tab 10 \tab 8 \tab 4 \tab 11 \tab 18 \tab 41 \tab - \tab 1 \tab 13 \tab - \tab 34 \tab 39 \tab 7 \tab - \tab - \tab 78 \tab 5 \tab 6 \tab - \tab 17 \tab - \tab 14 \tab 31 \tab 10 \tab 7 \tab 4 \tab 6 \tab 18 \tab 5 \tab 17 \tab 3 \tab - \tab 50 \tab 7 \tab 2 \tab 36 \tab 13 \tab 11 \tab 48 \tab 6 \tab 8 \tab 35 \tab 12 \tab 9 \tab 106 \tab 8 \tab 4 \tab 48 \tab 8 \tab 16 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 7 \tab 1 \tab - \tab 14 \tab 2 \tab 20 \tab - \tab 5 \tab 1 \tab 8 \tab - \tab 2 \tab 4 \tab - \tab 2 \tab 4 \tab 2 \tab 4 \tab 25 \tab 5 \tab 1 \tab 7 \tab 1 \tab 8 \tab - \tab 1 \tab 1 \tab 15 \tab 80 \tab 4 \tab - \tab 3 \tab 5 \tab 3 \tab 6 \tab 1 \tab 15 \tab - \tab 13 \tab 6 \tab 8 \tab 8 \tab 3 \tab 1 \tab 8 \tab 6 \tab 7 \tab 1 \tab - \tab 4 \tab 1 \tab 11 \tab 13 \tab 4 \tab 1 \tab 11 \tab 2 \tab 2 \tab 27 \tab 31 \tab 195 \tab - \tab 33 \tab 21 \tab 11 \tab 6 \tab 1 \tab 17 \tab 11 \tab 21 \tab 16 \tab 1 \tab 19 \tab 11 \tab 8 \tab 2 \tab 66 \tab 22 \tab 14 \tab 11 \tab 9 \tab 7 \tab 1 \tab 9 \tab 2 \tab 24 \tab 27 \tab 2 \tab 48 \tab 6 \tab - \tab 8 \tab 24 \tab 7 \tab 10 \tab 1 \tab 3 \tab 21 \tab 18 \tab 1 \tab - \tab 17 \tab - \tab - \tab - \tab 8 \tab 4 \tab 10 \tab - \tab 3 \tab 1 \tab - \tab 7 \tab - \tab 1 \tab 4 \tab 18 \tab 2 \tab 2 \tab 2 \tab 9 \tab - \tab 2 \tab - \tab - \tab 15 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 4 \tab 6 \tab 1 \tab - \tab - \tab - \tab 4 \tab 5 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 2 \tab 6 \tab - \tab 2 \tab 4 \tab 1 \tab 4 \tab - \tab - \tab 3 \tab 4 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 3 \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 164 \tab 23 \tab 1 \tab - \tab 1 \tab 3 \tab 25 \tab 23 \tab 2 \tab 11 \tab 9 \tab - \tab 2 \tab - \tab 7 \tab 3 \tab - \tab 5 \tab 2 \tab 3 \tab 18 \tab 9 \tab 3 \tab 33 \tab 1 \tab 2 \tab 24 \tab 2 \tab 3 \tab 55 \tab 28 \tab 16 \tab 6 \tab 34 \tab 57 \tab 22 \tab 5 \tab 3 \tab 1 \tab 10 \tab 2 \tab 1 \tab 3 \tab 2 \tab 6 \tab 8 \tab - \tab 10 \tab 24 \tab 45 \tab - \tab - \tab 22 \tab 2 \tab 38 \tab 35 \tab 10 \tab 1 \tab 1 \tab 82 \tab 7 \tab 10 \tab 2 \tab 18 \tab 1 \tab 20 \tab 20 \tab 11 \tab 2 \tab 5 \tab 13 \tab 2 \tab 2 \tab 10 \tab 2 \tab - \tab 49 \tab 8 \tab 1 \tab 34 \tab 12 \tab 12 \tab 30 \tab 4 \tab 9 \tab 29 \tab 21 \tab 16 \tab 84 \tab 7 \tab - \tab 42 \tab 10 \tab 19 \tab 3 \tab 7 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 16 \tab - \tab 17 \tab 1 \tab 1 \tab 1 \tab 7 \tab 2 \tab 4 \tab 2 \tab - \tab 4 \tab 4 \tab 10 \tab 2 \tab 24 \tab 5 \tab 2 \tab 3 \tab - \tab 4 \tab 1 \tab 5 \tab 6 \tab 18 \tab 77 \tab - \tab - \tab 3 \tab 5 \tab 2 \tab 4 \tab 1 \tab 21 \tab - \tab 8 \tab 7 \tab 9 \tab 11 \tab 2 \tab 1 \tab 5 \tab 11 \tab 16 \tab 2 \tab - \tab 3 \tab 5 \tab 22 \tab 8 \tab - \tab - \tab 9 \tab 5 \tab 2 \tab 17 \tab 27 \tab 210 \tab 3 \tab 14 \tab 25 \tab 10 \tab 5 \tab - \tab 9 \tab 10 \tab 22 \tab 27 \tab - \tab 29 \tab 8 \tab 4 \tab 3 \tab 62 \tab 16 \tab 21 \tab 4 \tab 6 \tab 14 \tab 3 \tab 5 \tab 3 \tab 9 \tab 40 \tab 9 \tab 65 \tab 4 \tab - \tab 2 \tab 27 \tab 9 \tab 9 \tab 1 \tab 2 \tab 11 \tab 35 \tab 2 \tab - \tab 21 \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 13 \tab - \tab - \tab 6 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 11 \tab 4 \tab 1 \tab 3 \tab 7 \tab 1 \tab 1 \tab - \tab 1 \tab 28 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 5 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 4 \tab 1 \tab 4 \tab 5 \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab 6 \tab 3 \tab 8 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 6 \tab 2 \tab - \tab 1 \tab 4 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 128 \tab 9 \tab 2 \tab 3 \tab 2 \tab 3 \tab 14 \tab 20 \tab - \tab 11 \tab 6 \tab 1 \tab 4 \tab - \tab 6 \tab 3 \tab - \tab 4 \tab - \tab 1 \tab 21 \tab 15 \tab 3 \tab 25 \tab 4 \tab 3 \tab 38 \tab 2 \tab 3 \tab 50 \tab 26 \tab 15 \tab 8 \tab 30 \tab 66 \tab 19 \tab 8 \tab 3 \tab - \tab 13 \tab 2 \tab 3 \tab 2 \tab 1 \tab 8 \tab 12 \tab 4 \tab 10 \tab 29 \tab 38 \tab 4 \tab - \tab 10 \tab - \tab 22 \tab 38 \tab 6 \tab 2 \tab 1 \tab 82 \tab 7 \tab 11 \tab 1 \tab 16 \tab - \tab 9 \tab 27 \tab 12 \tab 7 \tab 7 \tab 12 \tab 21 \tab 5 \tab 14 \tab 2 \tab 1 \tab 65 \tab 11 \tab 3 \tab 21 \tab 14 \tab 8 \tab 36 \tab 4 \tab 10 \tab 31 \tab 11 \tab 12 \tab 70 \tab 7 \tab - \tab 46 \tab 10 \tab 17 \tab 1 \tab 3 \tab 8 \tab 1 \tab - \tab - \tab 1 \tab 20 \tab - \tab 5 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab 14 \tab 1 \tab 2 \tab 3 \tab 8 \tab 3 \tab 5 \tab 7 \tab - \tab - \tab 4 \tab 3 \tab 6 \tab 30 \tab 6 \tab 2 \tab 11 \tab - \tab 9 \tab 1 \tab 2 \tab - \tab 12 \tab 71 \tab 1 \tab 1 \tab 5 \tab 2 \tab 9 \tab 5 \tab 3 \tab 18 \tab 3 \tab 10 \tab 4 \tab 13 \tab 5 \tab 3 \tab - \tab 5 \tab 6 \tab 17 \tab 6 \tab 1 \tab 1 \tab - \tab 19 \tab 13 \tab 1 \tab 1 \tab 15 \tab 6 \tab 7 \tab 21 \tab 31 \tab 171 \tab 2 \tab 34 \tab 20 \tab 7 \tab 7 \tab 3 \tab 11 \tab 9 \tab 10 \tab 24 \tab 5 \tab 25 \tab 10 \tab 5 \tab 6 \tab 48 \tab 17 \tab 6 \tab 9 \tab 4 \tab 18 \tab - \tab 10 \tab 2 \tab 12 \tab 26 \tab 7 \tab 60 \tab 1 \tab - \tab 5 \tab 32 \tab 7 \tab 6 \tab - \tab 2 \tab 21 \tab 35 \tab 1 \tab - \tab 31 \tab - \tab 3 \tab 2 \tab 7 \tab 6 \tab 6 \tab - \tab 4 \tab 8 \tab - \tab 3 \tab - \tab - \tab 3 \tab 12 \tab 1 \tab - \tab 2 \tab 11 \tab - \tab 1 \tab - \tab - \tab 15 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 3 \tab - \tab 5 \tab 3 \tab - \tab - \tab 1 \tab - \tab 6 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 7 \tab 3 \tab 3 \tab 8 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 9 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab 7 \tab 3 \tab - \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 163 \tab 13 \tab 2 \tab 1 \tab 1 \tab 3 \tab 15 \tab 14 \tab - \tab 5 \tab 5 \tab 3 \tab - \tab 1 \tab - \tab 7 \tab 1 \tab 4 \tab 1 \tab 3 \tab 18 \tab 12 \tab 3 \tab 23 \tab 3 \tab 1 \tab 18 \tab 1 \tab 2 \tab 47 \tab 27 \tab 21 \tab 4 \tab 27 \tab 69 \tab 24 \tab 7 \tab 1 \tab - \tab 15 \tab 3 \tab 8 \tab 3 \tab 3 \tab 7 \tab 11 \tab 2 \tab 3 \tab 19 \tab 40 \tab 1 \tab 2 \tab 8 \tab 1 \tab 33 \tab 43 \tab 10 \tab 2 \tab 2 \tab 69 \tab 7 \tab 14 \tab - \tab 10 \tab 1 \tab 18 \tab 31 \tab 12 \tab 4 \tab 2 \tab 11 \tab 11 \tab 4 \tab 17 \tab 3 \tab 2 \tab 48 \tab 19 \tab - \tab 24 \tab 5 \tab 7 \tab 27 \tab 5 \tab 11 \tab 40 \tab 10 \tab 11 \tab 61 \tab 6 \tab - \tab 45 \tab 13 \tab 11 \tab 2 \tab - \tab 8 \tab 3 \tab - \tab - \tab 1 \tab 9 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab 11 \tab - \tab 12 \tab - \tab 1 \tab 1 \tab 4 \tab 5 \tab 4 \tab 2 \tab - \tab - \tab 5 \tab 4 \tab 9 \tab 16 \tab 8 \tab 2 \tab 4 \tab - \tab 10 \tab 1 \tab 4 \tab 2 \tab 20 \tab 88 \tab - \tab - \tab 10 \tab 7 \tab 7 \tab 10 \tab 2 \tab 15 \tab 3 \tab 6 \tab 3 \tab 6 \tab 8 \tab 5 \tab 1 \tab 5 \tab 10 \tab 10 \tab 3 \tab 1 \tab - \tab - \tab 11 \tab 18 \tab 3 \tab 4 \tab 14 \tab 4 \tab 3 \tab 23 \tab 30 \tab 191 \tab 1 \tab 30 \tab 18 \tab 4 \tab 15 \tab 3 \tab 13 \tab 15 \tab 18 \tab 27 \tab 1 \tab 32 \tab 28 \tab 5 \tab 4 \tab 51 \tab 15 \tab 13 \tab 5 \tab 10 \tab 15 \tab 1 \tab 4 \tab 3 \tab 8 \tab 37 \tab 1 \tab 71 \tab 1 \tab - \tab 3 \tab 32 \tab 3 \tab 13 \tab 1 \tab 7 \tab 25 \tab 33 \tab 1 \tab 1 \tab 25 \tab - \tab - \tab 2 \tab 11 \tab 7 \tab 13 \tab 1 \tab 8 \tab 6 \tab - \tab 8 \tab - \tab - \tab 3 \tab 16 \tab 1 \tab 1 \tab 2 \tab 13 \tab - \tab - \tab - \tab 2 \tab 15 \tab 2 \tab - \tab - \tab 4 \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 4 \tab 5 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 5 \tab 7 \tab 2 \tab 5 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 9 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 268 \tab 13 \tab 3 \tab 5 \tab - \tab 5 \tab 26 \tab 48 \tab - \tab 17 \tab 7 \tab 3 \tab 2 \tab 3 \tab 2 \tab 10 \tab - \tab 9 \tab 8 \tab 2 \tab 39 \tab 19 \tab 8 \tab 45 \tab 2 \tab 5 \tab 63 \tab - \tab 5 \tab 71 \tab 46 \tab 22 \tab 9 \tab 58 \tab 72 \tab 23 \tab 12 \tab 5 \tab 2 \tab 18 \tab 6 \tab 3 \tab 5 \tab 1 \tab 5 \tab 24 \tab 8 \tab 17 \tab 42 \tab 67 \tab 3 \tab 3 \tab 12 \tab - \tab 44 \tab 58 \tab 14 \tab 5 \tab 1 \tab 127 \tab 7 \tab 9 \tab - \tab 19 \tab 1 \tab 27 \tab 48 \tab 9 \tab 20 \tab 3 \tab 9 \tab 26 \tab 4 \tab 37 \tab 8 \tab 3 \tab 98 \tab 11 \tab 3 \tab 40 \tab 7 \tab 23 \tab 56 \tab 9 \tab 11 \tab 55 \tab 29 \tab 34 \tab 120 \tab 10 \tab 3 \tab 63 \tab 22 \tab 21 \tab 3 \tab 12 \tab 5 \tab 2 \tab - \tab - \tab - \tab 9 \tab - \tab 5 \tab 1 \tab 4 \tab 1 \tab - \tab 10 \tab 3 \tab 1 \tab 10 \tab 3 \tab 21 \tab - \tab 3 \tab 16 \tab 8 \tab 1 \tab 9 \tab 5 \tab - \tab 1 \tab 8 \tab 5 \tab 11 \tab 30 \tab 12 \tab 4 \tab 7 \tab - \tab 40 \tab - \tab 10 \tab 1 \tab 30 \tab 87 \tab 8 \tab 7 \tab 18 \tab 6 \tab 9 \tab 9 \tab 4 \tab 46 \tab 3 \tab 15 \tab 8 \tab 12 \tab 15 \tab 5 \tab 2 \tab 1 \tab 13 \tab 11 \tab 3 \tab 2 \tab 10 \tab 2 \tab 21 \tab 28 \tab 1 \tab - \tab 14 \tab 7 \tab 7 \tab 26 \tab 57 \tab 250 \tab 1 \tab 38 \tab 37 \tab 7 \tab 19 \tab 2 \tab 23 \tab 34 \tab 18 \tab 61 \tab 3 \tab 44 \tab 14 \tab 7 \tab 6 \tab 80 \tab 28 \tab 12 \tab 8 \tab 11 \tab 10 \tab 1 \tab 14 \tab 4 \tab 28 \tab 38 \tab 8 \tab 87 \tab 5 \tab - \tab 4 \tab 84 \tab 12 \tab 12 \tab 1 \tab 2 \tab 30 \tab 50 \tab 6 \tab 2 \tab 23 \tab - \tab 3 \tab 4 \tab 13 \tab 12 \tab 14 \tab - \tab 3 \tab 17 \tab 1 \tab 16 \tab - \tab - \tab 6 \tab 28 \tab 3 \tab 8 \tab 2 \tab 20 \tab 2 \tab 1 \tab - \tab 7 \tab 41 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 5 \tab 6 \tab - \tab 5 \tab 7 \tab 2 \tab - \tab 5 \tab 1 \tab 9 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 8 \tab 6 \tab 6 \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab - \tab 4 \tab 2 \tab 9 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 12 \tab 8 \tab - \tab 3 \tab 3 \tab 7 \tab 1 \tab 5 \tab 2 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 9 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 5 \tab - \tab 4 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 112 \tab 8 \tab 1 \tab 1 \tab 1 \tab 5 \tab 14 \tab 14 \tab 3 \tab 5 \tab 4 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 17 \tab 14 \tab - \tab 9 \tab 1 \tab 2 \tab 28 \tab 1 \tab - \tab 30 \tab 30 \tab 13 \tab 5 \tab 28 \tab 39 \tab 16 \tab 10 \tab 4 \tab - \tab 8 \tab 1 \tab 4 \tab 1 \tab 1 \tab 5 \tab 13 \tab 15 \tab 6 \tab 19 \tab 40 \tab - \tab 2 \tab 5 \tab - \tab 27 \tab 36 \tab 4 \tab 2 \tab 1 \tab 51 \tab 3 \tab 6 \tab 1 \tab 16 \tab - \tab 11 \tab 27 \tab 5 \tab 3 \tab 2 \tab 3 \tab 15 \tab 2 \tab 6 \tab 3 \tab 1 \tab 37 \tab 4 \tab - \tab 21 \tab 7 \tab 5 \tab 27 \tab 5 \tab 8 \tab 17 \tab 9 \tab 7 \tab 40 \tab 19 \tab 2 \tab 39 \tab 9 \tab 12 \tab 1 \tab 4 \tab 2 \tab 3 \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab 7 \tab - \tab 5 \tab - \tab 1 \tab 6 \tab 5 \tab 5 \tab 8 \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 12 \tab 3 \tab 2 \tab 6 \tab - \tab 10 \tab - \tab 1 \tab - \tab 17 \tab 30 \tab 2 \tab - \tab 11 \tab 4 \tab 3 \tab 3 \tab 2 \tab 3 \tab 1 \tab 6 \tab 7 \tab 3 \tab 3 \tab - \tab 1 \tab 4 \tab 11 \tab 7 \tab 2 \tab 1 \tab - \tab - \tab 9 \tab 16 \tab - \tab 2 \tab 10 \tab 6 \tab 3 \tab 25 \tab 20 \tab 138 \tab 1 \tab 15 \tab 12 \tab 2 \tab 7 \tab 2 \tab 8 \tab 11 \tab 15 \tab 5 \tab 2 \tab 19 \tab 9 \tab 3 \tab 3 \tab 36 \tab 20 \tab 9 \tab 5 \tab 7 \tab 5 \tab 2 \tab 7 \tab 2 \tab 13 \tab 25 \tab 9 \tab 35 \tab - \tab - \tab 1 \tab 22 \tab 7 \tab 8 \tab 1 \tab 2 \tab 14 \tab 16 \tab 1 \tab - \tab 9 \tab - \tab - \tab 1 \tab 3 \tab 5 \tab 4 \tab 1 \tab 2 \tab 7 \tab - \tab 8 \tab - \tab - \tab - \tab 12 \tab 1 \tab 4 \tab 2 \tab 5 \tab - \tab 2 \tab 1 \tab 1 \tab 23 \tab - \tab 4 \tab - \tab 2 \tab - \tab - \tab 2 \tab 6 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 4 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab 169 \tab 11 \tab 2 \tab 1 \tab 1 \tab 1 \tab 10 \tab 17 \tab - \tab 11 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab - \tab 1 \tab 18 \tab 8 \tab 2 \tab 15 \tab 1 \tab - \tab 29 \tab - \tab 3 \tab 46 \tab 18 \tab 8 \tab 4 \tab 24 \tab 35 \tab 8 \tab 4 \tab 1 \tab - \tab 8 \tab 1 \tab 5 \tab 1 \tab 1 \tab 3 \tab 5 \tab 20 \tab 8 \tab 23 \tab 28 \tab 2 \tab - \tab 6 \tab 2 \tab 27 \tab 40 \tab 3 \tab 3 \tab 1 \tab 61 \tab 3 \tab 6 \tab - \tab 6 \tab - \tab 8 \tab 20 \tab 6 \tab 6 \tab 3 \tab 4 \tab 11 \tab 2 \tab 5 \tab - \tab 3 \tab 27 \tab 6 \tab - \tab 18 \tab 6 \tab 3 \tab 27 \tab 1 \tab 7 \tab 19 \tab 12 \tab 7 \tab 52 \tab 22 \tab 2 \tab 51 \tab 7 \tab 8 \tab 3 \tab 6 \tab 6 \tab - \tab - \tab 1 \tab 1 \tab 10 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab 5 \tab 1 \tab 7 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 3 \tab 3 \tab - \tab 2 \tab 2 \tab 2 \tab 3 \tab 13 \tab 1 \tab 2 \tab 3 \tab - \tab 11 \tab 1 \tab 3 \tab - \tab 35 \tab 53 \tab 2 \tab 3 \tab 10 \tab 1 \tab 2 \tab 2 \tab 2 \tab 11 \tab - \tab 7 \tab 5 \tab 2 \tab 3 \tab 1 \tab - \tab 5 \tab 13 \tab 3 \tab - \tab 1 \tab 3 \tab - \tab 4 \tab 9 \tab - \tab - \tab 11 \tab 3 \tab 3 \tab 15 \tab 25 \tab 140 \tab 2 \tab 13 \tab 12 \tab 4 \tab 12 \tab 1 \tab 11 \tab 20 \tab 12 \tab 2 \tab 1 \tab 13 \tab 13 \tab 7 \tab 3 \tab 37 \tab 14 \tab 8 \tab 5 \tab 6 \tab 6 \tab 3 \tab 5 \tab 1 \tab 4 \tab 16 \tab 2 \tab 47 \tab - \tab - \tab 1 \tab 22 \tab 13 \tab 5 \tab - \tab - \tab 11 \tab 12 \tab 3 \tab - \tab 12 \tab - \tab 1 \tab - \tab 8 \tab 3 \tab 6 \tab - \tab 2 \tab 7 \tab 2 \tab 10 \tab 1 \tab 1 \tab 3 \tab 10 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab 11 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 6 \tab 3 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 4 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 41113 \tab 227 \tab 32 \tab 32 \tab 19 \tab 33 \tab 198 \tab 242 \tab 29 \tab 131 \tab 73 \tab 16 \tab 32 \tab 9 \tab 43 \tab 55 \tab 17 \tab 53 \tab 42 \tab 15 \tab 284 \tab 146 \tab 36 \tab 332 \tab 20 \tab 32 \tab 385 \tab 13 \tab 38 \tab 543 \tab 346 \tab 160 \tab 78 \tab 416 \tab 672 \tab 243 \tab 85 \tab 37 \tab 10 \tab 146 \tab 28 \tab 60 \tab 25 \tab 13 \tab 91 \tab 143 \tab 65 \tab 98 \tab 303 \tab 485 \tab 13 \tab 23 \tab 175 \tab 20 \tab 376 \tab 466 \tab 90 \tab 19 \tab 10 \tab 974 \tab 57 \tab 101 \tab 5 \tab 181 \tab 8 \tab 211 \tab 320 \tab 123 \tab 66 \tab 67 \tab 86 \tab 173 \tab 49 \tab 186 \tab 47 \tab 27 \tab 630 \tab 110 \tab 22 \tab 397 \tab 117 \tab 120 \tab 414 \tab 66 \tab 103 \tab 429 \tab 181 \tab 144 \tab 1133 \tab 107 \tab 23 \tab 523 \tab 133 \tab 173 \tab 32 \tab 49 \tab 65 \tab 22 \tab 4 \tab 3 \tab 15 \tab 74 \tab 1 \tab 35 \tab 14 \tab 21 \tab 9 \tab 1 \tab 45 \tab 14 \tab 11 \tab 161 \tab 14 \tab 158 \tab 8 \tab 32 \tab 51 \tab 75 \tab 21 \tab 68 \tab 47 \tab 6 \tab 19 \tab 66 \tab 42 \tab 58 \tab 255 \tab 66 \tab 30 \tab 80 \tab 6 \tab 143 \tab 5 \tab 38 \tab 21 \tab 229 \tab 831 \tab 48 \tab 20 \tab 145 \tab 58 \tab 56 \tab 69 \tab 35 \tab 201 \tab 21 \tab 115 \tab 77 \tab 89 \tab 87 \tab 35 \tab 13 \tab 56 \tab 110 \tab 116 \tab 38 \tab 10 \tab 31 \tab 10 \tab 144 \tab 159 \tab 16 \tab 16 \tab 139 \tab 60 \tab 57 \tab 247 \tab 358 \tab 1958 \tab 12 \tab 285 \tab 218 \tab 79 \tab 120 \tab 19 \tab 160 \tab 163 \tab 186 \tab 190 \tab 34 \tab 276 \tab 151 \tab 62 \tab 47 \tab 588 \tab 182 \tab 131 \tab 72 \tab 81 \tab 126 \tab 26 \tab 83 \tab 33 \tab 145 \tab 351 \tab 58 \tab 695 \tab 25 \tab 2 \tab 37 \tab 361 \tab 102 \tab 86 \tab 11 \tab 31 \tab 176 \tab 268 \tab 26 \tab 7 \tab 197 \tab 4 \tab 10 \tab 11 \tab 76 \tab 68 \tab 113 \tab 8 \tab 36 \tab 64 \tab 9 \tab 87 \tab 4 \tab 8 \tab 35 \tab 176 \tab 19 \tab 28 \tab 28 \tab 115 \tab 7 \tab 17 \tab 4 \tab 20 \tab 188 \tab 12 \tab 10 \tab 5 \tab 22 \tab 8 \tab 3 \tab 44 \tab 34 \tab 5 \tab 36 \tab 45 \tab 6 \tab 4 \tab 13 \tab 17 \tab 51 \tab 25 \tab 30 \tab 3 \tab 20 \tab 11 \tab 18 \tab 11 \tab 16 \tab 36 \tab 37 \tab 37 \tab 67 \tab 4 \tab 14 \tab 20 \tab 5 \tab 22 \tab 4 \tab 2 \tab 17 \tab 42 \tab 19 \tab 62 \tab 6 \tab 3 \tab 7 \tab 2 \tab 5 \tab 2 \tab 15 \tab 2 \tab 37 \tab 38 \tab 18 \tab 14 \tab 11 \tab 31 \tab 17 \tab 29 \tab 14 \tab 11 \tab 15 \tab 2 \tab 4 \tab 6 \tab 2 \tab 14 \tab 5 \tab 16 \tab 31 \tab 6 \tab 3 \tab 1 \tab 1 \tab 9 \tab 11 \tab 3 \tab 7 \tab 6 \tab 6 \tab 8 \tab 3 \tab 7 \tab 4 \tab 2 \tab 22 \tab 3 \tab 7 \tab 8 \tab 4 \tab 7 \tab 2 \tab 2 \tab 8 \tab 3 \tab 4 \tab 3 \tab 6 \tab 3 \tab 2 \tab 1 \tab 5 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} i don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
