% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_597.R
\name{sei10}
\alias{sei10}
\title{R's socioeconomic index (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sei10
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 10.6 \tab 11.8 \tab 12.4 \tab 12.6 \tab 12.7 \tab 13.2 \tab 13.3 \tab 13.6 \tab 14 \tab 14.4 \tab 14.6 \tab 14.8 \tab 15.8 \tab 17.1 \tab 17.5 \tab 18.6 \tab 18.8 \tab 19 \tab 19.2 \tab 19.6 \tab 19.7 \tab 20 \tab 20.3 \tab 20.5 \tab 20.7 \tab 20.8 \tab 20.9 \tab 21 \tab 21.1 \tab 21.4 \tab 21.6 \tab 21.8 \tab 22.1 \tab 22.4 \tab 22.9 \tab 23 \tab 23.2 \tab 23.3 \tab 23.7 \tab 23.8 \tab 24 \tab 24.1 \tab 24.2 \tab 24.6 \tab 25 \tab 25.1 \tab 25.2 \tab 25.7 \tab 26.1 \tab 26.2 \tab 26.5 \tab 26.6 \tab 26.8 \tab 26.9 \tab 27 \tab 27.1 \tab 27.2 \tab 27.5 \tab 27.7 \tab 28.3 \tab 28.4 \tab 28.6 \tab 28.7 \tab 29.2 \tab 29.5 \tab 29.7 \tab 29.9 \tab 30.3 \tab 31 \tab 31.1 \tab 31.2 \tab 31.3 \tab 31.5 \tab 31.6 \tab 31.9 \tab 32 \tab 32.6 \tab 32.7 \tab 33.4 \tab 33.7 \tab 34.1 \tab 34.7 \tab 34.8 \tab 35.1 \tab 35.2 \tab 35.3 \tab 35.8 \tab 36.2 \tab 36.4 \tab 36.5 \tab 37 \tab 37.2 \tab 37.3 \tab 37.4 \tab 37.6 \tab 37.7 \tab 38.1 \tab 38.2 \tab 38.3 \tab 38.8 \tab 38.9 \tab 39.1 \tab 39.2 \tab 39.6 \tab 39.7 \tab 39.9 \tab 40.1 \tab 41 \tab 41.1 \tab 41.8 \tab 42.8 \tab 42.9 \tab 43 \tab 43.8 \tab 44.9 \tab 45 \tab 45.3 \tab 45.4 \tab 46 \tab 46.5 \tab 46.6 \tab 47.9 \tab 48.1 \tab 48.3 \tab 48.8 \tab 49.4 \tab 49.6 \tab 50 \tab 50.3 \tab 50.4 \tab 52 \tab 52.1 \tab 52.4 \tab 52.5 \tab 52.6 \tab 52.9 \tab 53.5 \tab 53.6 \tab 53.8 \tab 54.6 \tab 55 \tab 55.1 \tab 56.8 \tab 56.9 \tab 57.1 \tab 57.4 \tab 57.5 \tab 57.8 \tab 58.1 \tab 58.2 \tab 58.4 \tab 59.1 \tab 59.8 \tab 60.1 \tab 60.4 \tab 60.5 \tab 61.2 \tab 61.4 \tab 62 \tab 62.4 \tab 62.6 \tab 62.7 \tab 62.9 \tab 63.7 \tab 64.2 \tab 64.9 \tab 65.1 \tab 65.2 \tab 65.3 \tab 66.4 \tab 67.7 \tab 67.8 \tab 68 \tab 68.6 \tab 69.3 \tab 69.4 \tab 70.3 \tab 71 \tab 71.6 \tab 71.7 \tab 71.8 \tab 72.7 \tab 73.4 \tab 73.6 \tab 73.9 \tab 74.6 \tab 75.5 \tab 76.3 \tab 76.7 \tab 77.2 \tab 77.4 \tab 80.7 \tab 80.9 \tab 81 \tab 81.6 \tab 81.9 \tab 82.4 \tab 82.5 \tab 83.8 \tab 84 \tab 84.2 \tab 84.5 \tab 85.2 \tab 86.5 \tab 87.6 \tab 87.9 \tab 88.7 \tab 89.3 \tab 89.9 \tab 9 \tab 91.1 \tab 91.9 \tab 92.1 \tab 92.8 \tab iap \tab 17 \tab 20.1 \tab 24.8 \tab 25.5 \tab 27.4 \tab 28.5 \tab 29.4 \tab 30.1 \tab 30.9 \tab 33.9 \tab 34.4 \tab 34.5 \tab 34.6 \tab 36.7 \tab 39 \tab 39.5 \tab 39.8 \tab 40 \tab 41.6 \tab 42.2 \tab 43.4 \tab 43.5 \tab 45.1 \tab 46.2 \tab 49.2 \tab 50.1 \tab 50.7 \tab 51.7 \tab 51.8 \tab 52.3 \tab 53 \tab 53.2 \tab 53.9 \tab 54.3 \tab 55.3 \tab 57.3 \tab 58.6 \tab 59.9 \tab 60.2 \tab 64.5 \tab 66.2 \tab 70.1 \tab 73.1 \tab 78.3 \tab 79.3 \tab 82.9 \tab 84.3 \tab 85.5 \tab 86.1 \tab 86.6 \tab 86.7 \tab 88.3 \tab 89.1 \tab 91.6 \tab 22.5 \tab 30 \tab 31.4 \tab 33.2 \tab 35.4 \tab 37.1 \tab 41.2 \tab 44.6 \tab 51.5 \tab 51.9 \tab 60.3 \tab 62.1 \tab 63.1 \tab 65.6 \tab 70.5 \tab 72.5 \tab 77.3 \tab 77.9 \tab 87.3 \tab 93.7 \tab 21.5 \tab 25.6 \tab 26.7 \tab 27.6 \tab 32.8 \tab 33.1 \tab 44.5 \tab 49.3 \tab 59.4 \tab 61.5 \tab 63.2 \tab 70.4 \tab 81.5 \tab 82.7 \tab 85 \tab 92 \tab 18.9 \tab 21.3 \tab 25.4 \tab 29.6 \tab 30.8 \tab 33.3 \tab 35.5 \tab 43.1 \tab 48.4 \tab 53.7 \tab 59 \tab 63.3 \tab 79.9 \tab 85.8 \tab 85.9 \tab 88.6 \tab 28.8 \tab 36.9 \tab 41.4 \tab 57 \tab 75.6 \tab 77.7 \tab 79.4 \tab 28.1 \tab 37.8 \tab 42.7 \tab 53.1 \tab 54.1 \tab 66.1 \tab 73.5 \tab 75.2 \tab 75.7 \tab 92.2 \tab 21.2 \tab 29.3 \tab 36.8 \tab 37.9 \tab 71.4 \tab 80.1 \tab 86.8 \tab 89.4 \tab 32.1 \tab 38.4 \tab 64.4 \tab 77 \tab 80.2 \tab 83.4 \tab 40.4 \tab 71.1 \tab 73 \tab 74.8 \tab 72.6 \tab 73.2 \tab 83.1 \tab 46.7 \tab 88 \tab 76.5 \tab 87.8 \tab 88.4 \tab 51.6 \tab 56.3 \tab 69.5 \tab 79.6 \tab 61.9 \tab 86.9 \tab 42.3 \tab 86 \tab Total \cr
   1972 \tab 1 \tab 8 \tab 4 \tab 53 \tab 1 \tab 19 \tab 1 \tab 23 \tab 14 \tab 1 \tab 3 \tab 11 \tab 5 \tab 3 \tab 1 \tab 7 \tab 5 \tab 1 \tab 1 \tab 27 \tab 3 \tab 4 \tab 1 \tab 3 \tab 25 \tab 7 \tab 11 \tab 7 \tab 1 \tab 1 \tab 18 \tab 8 \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab 21 \tab 5 \tab 4 \tab 6 \tab 3 \tab 14 \tab 11 \tab 2 \tab 3 \tab 10 \tab 5 \tab 6 \tab 4 \tab 6 \tab 1 \tab 7 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 14 \tab 5 \tab 12 \tab 2 \tab 4 \tab 2 \tab 3 \tab 1 \tab 8 \tab 8 \tab 4 \tab 1 \tab 1 \tab 5 \tab 13 \tab 48 \tab 8 \tab 1 \tab 1 \tab 1 \tab 7 \tab 4 \tab 1 \tab 14 \tab 3 \tab 9 \tab 29 \tab 9 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 16 \tab 1 \tab 34 \tab 74 \tab 1 \tab 1 \tab 2 \tab 2 \tab 19 \tab 41 \tab 4 \tab 7 \tab 7 \tab 4 \tab 3 \tab 2 \tab 11 \tab 2 \tab 2 \tab 1 \tab 6 \tab 9 \tab 6 \tab 26 \tab 10 \tab 3 \tab 1 \tab 4 \tab 2 \tab 2 \tab 2 \tab 13 \tab 1 \tab 16 \tab 11 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 8 \tab 2 \tab 5 \tab 3 \tab 1 \tab 9 \tab 8 \tab 5 \tab 2 \tab 3 \tab 2 \tab 2 \tab 3 \tab 12 \tab 1 \tab 6 \tab 4 \tab 3 \tab 3 \tab 1 \tab 1 \tab 4 \tab 1 \tab 11 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 8 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 3 \tab 6 \tab 8 \tab 1 \tab 1 \tab 1 \tab 2 \tab 31 \tab 3 \tab 2 \tab 1 \tab 1 \tab 19 \tab 3 \tab 1 \tab 23 \tab 11 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 334 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 3 \tab 1 \tab 3 \tab 30 \tab 3 \tab 18 \tab 1 \tab 16 \tab 20 \tab 4 \tab 1 \tab 9 \tab 5 \tab 5 \tab - \tab 4 \tab 4 \tab - \tab 2 \tab 24 \tab 4 \tab 2 \tab - \tab 2 \tab 24 \tab 5 \tab 4 \tab 5 \tab 1 \tab - \tab 24 \tab 10 \tab 2 \tab - \tab 6 \tab - \tab 2 \tab 11 \tab 6 \tab 13 \tab 3 \tab 4 \tab 23 \tab 15 \tab 13 \tab 8 \tab 10 \tab 8 \tab 5 \tab 3 \tab 10 \tab - \tab 17 \tab 2 \tab 5 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 6 \tab 5 \tab 22 \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab 3 \tab 4 \tab 5 \tab 1 \tab - \tab 4 \tab 6 \tab 30 \tab 10 \tab 1 \tab 2 \tab 1 \tab 4 \tab 8 \tab 1 \tab 9 \tab 2 \tab 2 \tab 18 \tab 6 \tab 2 \tab 2 \tab - \tab 4 \tab 2 \tab - \tab 1 \tab 1 \tab 18 \tab - \tab 48 \tab 87 \tab - \tab - \tab 2 \tab 2 \tab 36 \tab 20 \tab 1 \tab 5 \tab 1 \tab 2 \tab 5 \tab 1 \tab 10 \tab 1 \tab - \tab - \tab 5 \tab 6 \tab 7 \tab 16 \tab 5 \tab 2 \tab - \tab 2 \tab 3 \tab 5 \tab 3 \tab 3 \tab 2 \tab 7 \tab 11 \tab 1 \tab 8 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 7 \tab 3 \tab 10 \tab 2 \tab 1 \tab 4 \tab 7 \tab 7 \tab 1 \tab 1 \tab 2 \tab 2 \tab 4 \tab 18 \tab - \tab 2 \tab 3 \tab 4 \tab 3 \tab - \tab - \tab 4 \tab 2 \tab 5 \tab 1 \tab 6 \tab - \tab 2 \tab 3 \tab 9 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 6 \tab - \tab 7 \tab 5 \tab - \tab 1 \tab 6 \tab 3 \tab 26 \tab 2 \tab 2 \tab 2 \tab - \tab 23 \tab 5 \tab - \tab 13 \tab 14 \tab 7 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 5 \tab - \tab 1 \tab 1 \tab 245 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 5 \tab 2 \tab 1 \tab 31 \tab 3 \tab 19 \tab 3 \tab 18 \tab 9 \tab 2 \tab 2 \tab 7 \tab 5 \tab 5 \tab - \tab 6 \tab 1 \tab 2 \tab - \tab 24 \tab - \tab 2 \tab - \tab 3 \tab 21 \tab 8 \tab 3 \tab - \tab 1 \tab 1 \tab 24 \tab 7 \tab 1 \tab - \tab 6 \tab 1 \tab 2 \tab 25 \tab 4 \tab 13 \tab 2 \tab 1 \tab 25 \tab 12 \tab 5 \tab 2 \tab 8 \tab 5 \tab 8 \tab 2 \tab 5 \tab - \tab 21 \tab 3 \tab 4 \tab 2 \tab 1 \tab 2 \tab 4 \tab 1 \tab 7 \tab 7 \tab 20 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 6 \tab 6 \tab 9 \tab 2 \tab 1 \tab 8 \tab 8 \tab 23 \tab 11 \tab 2 \tab 7 \tab 1 \tab 4 \tab 8 \tab 1 \tab 21 \tab 1 \tab 2 \tab 12 \tab 13 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab 17 \tab 1 \tab 32 \tab 73 \tab 1 \tab - \tab 2 \tab 2 \tab 27 \tab 24 \tab 1 \tab 4 \tab 6 \tab 1 \tab 5 \tab 2 \tab 22 \tab 3 \tab 2 \tab - \tab 6 \tab 4 \tab 5 \tab 20 \tab 11 \tab 4 \tab - \tab 1 \tab 2 \tab 2 \tab 3 \tab 6 \tab 1 \tab 10 \tab 12 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 2 \tab 3 \tab 9 \tab 1 \tab 2 \tab 4 \tab 3 \tab 8 \tab 4 \tab 3 \tab 3 \tab 1 \tab 5 \tab 11 \tab 1 \tab 5 \tab 4 \tab 6 \tab 3 \tab - \tab - \tab 1 \tab - \tab 12 \tab 1 \tab 7 \tab 1 \tab 3 \tab 5 \tab 8 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab 10 \tab 5 \tab 1 \tab 3 \tab 3 \tab - \tab 40 \tab - \tab 3 \tab - \tab 1 \tab 30 \tab 2 \tab 1 \tab 17 \tab 11 \tab - \tab 5 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 6 \tab - \tab - \tab 3 \tab 225 \tab 2 \tab 2 \tab - \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 4 \tab 3 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 2 \tab 5 \tab 4 \tab 38 \tab 2 \tab 38 \tab 2 \tab 17 \tab 15 \tab 3 \tab 2 \tab 8 \tab 1 \tab 6 \tab 1 \tab 4 \tab 8 \tab - \tab 1 \tab 29 \tab 3 \tab 4 \tab - \tab 1 \tab 23 \tab 7 \tab 3 \tab 3 \tab 2 \tab 3 \tab 19 \tab 11 \tab 1 \tab - \tab 5 \tab - \tab 3 \tab 19 \tab 5 \tab 15 \tab 6 \tab 1 \tab 23 \tab 18 \tab 6 \tab 5 \tab 10 \tab 12 \tab 9 \tab 1 \tab 8 \tab 1 \tab 9 \tab 6 \tab 4 \tab - \tab 1 \tab 2 \tab 3 \tab 4 \tab 8 \tab 7 \tab 21 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 4 \tab 13 \tab - \tab 3 \tab 8 \tab 4 \tab 29 \tab 6 \tab - \tab 5 \tab - \tab 1 \tab 4 \tab - \tab 10 \tab 1 \tab 4 \tab 13 \tab 7 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 15 \tab 1 \tab 44 \tab 80 \tab - \tab 3 \tab 2 \tab 1 \tab 21 \tab 20 \tab - \tab 5 \tab 4 \tab 3 \tab 4 \tab - \tab 18 \tab 1 \tab 1 \tab - \tab 8 \tab 9 \tab 11 \tab 14 \tab 17 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 5 \tab 1 \tab 6 \tab 11 \tab 3 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 4 \tab 7 \tab 1 \tab 1 \tab 8 \tab 6 \tab 6 \tab 3 \tab - \tab 6 \tab 1 \tab 8 \tab 11 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 8 \tab - \tab 6 \tab 1 \tab 1 \tab 2 \tab 11 \tab - \tab - \tab 1 \tab - \tab - \tab 6 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 4 \tab 9 \tab - \tab 2 \tab 2 \tab 1 \tab 27 \tab 5 \tab 2 \tab 1 \tab - \tab 15 \tab 4 \tab 1 \tab 17 \tab 5 \tab 1 \tab 8 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 246 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab 6 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 3 \tab 4 \tab 2 \tab 36 \tab 2 \tab 24 \tab 1 \tab 23 \tab 15 \tab 1 \tab 1 \tab 8 \tab 7 \tab 2 \tab 2 \tab 10 \tab 1 \tab - \tab - \tab 27 \tab 4 \tab 1 \tab 1 \tab 3 \tab 19 \tab 2 \tab 2 \tab 4 \tab 2 \tab 1 \tab 21 \tab 6 \tab - \tab - \tab 7 \tab 1 \tab - \tab 23 \tab 14 \tab 11 \tab 5 \tab 1 \tab 16 \tab 9 \tab 12 \tab 6 \tab 6 \tab 10 \tab 10 \tab 3 \tab 7 \tab - \tab 26 \tab 4 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 12 \tab 1 \tab 6 \tab - \tab - \tab - \tab 1 \tab 4 \tab 4 \tab 4 \tab 9 \tab 3 \tab 1 \tab 4 \tab 14 \tab 29 \tab 17 \tab 2 \tab 3 \tab 5 \tab 1 \tab 3 \tab - \tab 15 \tab 1 \tab 12 \tab 18 \tab 5 \tab 4 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 13 \tab - \tab 40 \tab 72 \tab - \tab 3 \tab 2 \tab 1 \tab 32 \tab 35 \tab 1 \tab 6 \tab 3 \tab 2 \tab 5 \tab 2 \tab 26 \tab - \tab - \tab - \tab 5 \tab 4 \tab 4 \tab 23 \tab 12 \tab 2 \tab - \tab 2 \tab 3 \tab 5 \tab 1 \tab 5 \tab 1 \tab 5 \tab 14 \tab 2 \tab - \tab 4 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 6 \tab - \tab 9 \tab 4 \tab 11 \tab - \tab 1 \tab 1 \tab 5 \tab 4 \tab 3 \tab 1 \tab 2 \tab - \tab 2 \tab 9 \tab 1 \tab 7 \tab 4 \tab 6 \tab 4 \tab 2 \tab - \tab 4 \tab 1 \tab 10 \tab 1 \tab 9 \tab 2 \tab 4 \tab 8 \tab 8 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 6 \tab - \tab 4 \tab 7 \tab - \tab 3 \tab 9 \tab 2 \tab 28 \tab 4 \tab 1 \tab - \tab 1 \tab 22 \tab 4 \tab 1 \tab 15 \tab 10 \tab 3 \tab 3 \tab 1 \tab - \tab 4 \tab - \tab - \tab 2 \tab 5 \tab 1 \tab - \tab 2 \tab 196 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1 \tab 1 \tab 5 \tab 31 \tab 4 \tab 24 \tab 5 \tab 14 \tab 14 \tab - \tab 5 \tab 15 \tab 3 \tab 10 \tab 3 \tab 4 \tab 5 \tab - \tab 1 \tab 30 \tab 5 \tab 3 \tab 3 \tab 1 \tab 20 \tab 9 \tab 3 \tab 4 \tab 2 \tab - \tab 22 \tab 6 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 21 \tab 5 \tab 10 \tab 10 \tab 4 \tab 20 \tab 16 \tab 13 \tab - \tab 15 \tab 6 \tab 9 \tab 6 \tab 5 \tab - \tab 18 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 5 \tab 2 \tab 15 \tab 1 \tab 22 \tab 2 \tab 1 \tab - \tab 4 \tab - \tab 7 \tab 3 \tab 13 \tab 1 \tab - \tab 5 \tab 8 \tab 32 \tab 5 \tab - \tab 7 \tab - \tab 2 \tab 6 \tab 1 \tab 6 \tab 4 \tab 5 \tab 21 \tab 4 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 14 \tab 2 \tab 30 \tab 74 \tab 1 \tab 1 \tab 3 \tab - \tab 21 \tab 25 \tab - \tab 7 \tab 5 \tab 4 \tab 4 \tab 1 \tab 14 \tab 3 \tab 3 \tab - \tab 5 \tab 7 \tab 5 \tab 15 \tab 11 \tab 3 \tab - \tab 1 \tab 2 \tab 5 \tab 1 \tab 3 \tab 1 \tab 11 \tab 10 \tab 4 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 10 \tab - \tab 7 \tab 4 \tab 4 \tab - \tab 1 \tab 3 \tab 9 \tab 5 \tab 1 \tab 1 \tab 4 \tab 1 \tab 7 \tab 5 \tab - \tab 8 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 7 \tab - \tab 9 \tab 1 \tab - \tab 3 \tab 9 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 5 \tab 10 \tab - \tab 2 \tab 4 \tab 3 \tab 35 \tab 1 \tab 4 \tab - \tab - \tab 27 \tab 2 \tab 1 \tab 18 \tab 11 \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab 3 \tab 2 \tab 270 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 7 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 3 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 3 \tab 5 \tab 5 \tab 30 \tab 1 \tab 31 \tab 2 \tab 21 \tab 17 \tab 1 \tab 1 \tab 7 \tab 4 \tab 11 \tab - \tab 11 \tab 3 \tab - \tab 2 \tab 33 \tab 2 \tab 3 \tab - \tab 4 \tab 18 \tab 7 \tab 3 \tab 2 \tab 2 \tab - \tab 15 \tab 6 \tab 2 \tab - \tab 4 \tab 2 \tab 4 \tab 28 \tab 8 \tab 8 \tab 3 \tab 2 \tab 21 \tab 16 \tab 11 \tab 9 \tab 5 \tab 6 \tab 4 \tab 4 \tab 8 \tab 1 \tab 22 \tab 4 \tab 7 \tab - \tab 1 \tab 2 \tab 5 \tab 2 \tab 10 \tab 5 \tab 22 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 4 \tab 6 \tab 5 \tab 4 \tab - \tab 7 \tab 5 \tab 31 \tab 11 \tab - \tab 5 \tab 1 \tab 1 \tab 6 \tab 1 \tab 13 \tab - \tab 4 \tab 14 \tab 2 \tab 3 \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab 5 \tab - \tab 21 \tab 2 \tab 49 \tab 93 \tab - \tab 3 \tab 3 \tab 1 \tab 34 \tab 16 \tab 1 \tab 6 \tab 1 \tab 9 \tab 9 \tab 2 \tab 23 \tab 1 \tab - \tab 1 \tab 4 \tab 10 \tab 6 \tab 20 \tab 12 \tab 1 \tab - \tab 4 \tab - \tab 2 \tab 2 \tab 8 \tab 1 \tab 8 \tab 17 \tab 7 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 8 \tab - \tab 6 \tab 11 \tab 9 \tab 1 \tab 1 \tab 2 \tab 9 \tab 3 \tab 3 \tab 1 \tab 5 \tab 2 \tab 2 \tab 13 \tab - \tab 4 \tab 3 \tab 4 \tab 3 \tab 2 \tab - \tab 4 \tab 1 \tab 10 \tab 2 \tab 5 \tab 2 \tab 2 \tab 5 \tab 3 \tab - \tab - \tab 3 \tab 2 \tab 6 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 7 \tab 9 \tab 3 \tab 1 \tab 5 \tab - \tab 32 \tab 3 \tab 1 \tab - \tab 1 \tab 22 \tab 4 \tab 4 \tab 19 \tab 7 \tab 4 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 5 \tab 131 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 6 \tab 1 \tab 2 \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 3 \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 2 \tab - \tab - \tab 28 \tab 2 \tab 24 \tab 3 \tab 20 \tab 15 \tab 4 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab 5 \tab 9 \tab 1 \tab 2 \tab 27 \tab 3 \tab 1 \tab 1 \tab 1 \tab 22 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab 13 \tab 10 \tab 3 \tab - \tab 9 \tab - \tab 3 \tab 19 \tab 3 \tab 7 \tab 7 \tab 1 \tab 24 \tab 13 \tab 8 \tab 3 \tab 6 \tab 12 \tab 6 \tab 2 \tab 2 \tab - \tab 15 \tab 9 \tab 6 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 12 \tab 3 \tab 14 \tab - \tab - \tab - \tab 2 \tab - \tab 5 \tab 6 \tab 9 \tab 2 \tab - \tab 4 \tab 7 \tab 30 \tab 13 \tab 1 \tab 4 \tab 2 \tab 2 \tab 4 \tab - \tab 6 \tab - \tab 9 \tab 13 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 3 \tab - \tab 12 \tab 6 \tab 40 \tab 80 \tab 2 \tab 1 \tab 3 \tab - \tab 28 \tab 28 \tab 1 \tab 11 \tab 5 \tab 3 \tab 7 \tab - \tab 27 \tab 1 \tab 1 \tab - \tab 7 \tab 5 \tab 8 \tab 17 \tab 12 \tab 3 \tab - \tab 4 \tab 2 \tab 3 \tab 1 \tab 4 \tab 2 \tab 5 \tab 19 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 3 \tab 9 \tab - \tab 8 \tab 6 \tab 7 \tab - \tab 1 \tab - \tab 6 \tab 7 \tab 3 \tab 3 \tab 3 \tab - \tab 4 \tab 7 \tab 4 \tab - \tab 5 \tab 5 \tab 4 \tab 1 \tab - \tab 3 \tab - \tab 9 \tab - \tab 22 \tab - \tab 3 \tab 4 \tab 8 \tab - \tab - \tab 4 \tab - \tab - \tab 3 \tab - \tab 3 \tab - \tab 3 \tab 2 \tab 7 \tab 10 \tab - \tab 2 \tab 6 \tab 1 \tab 32 \tab 10 \tab 2 \tab 1 \tab 1 \tab 17 \tab 2 \tab 1 \tab 18 \tab 8 \tab 2 \tab 6 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 8 \tab - \tab - \tab 3 \tab 167 \tab 1 \tab 2 \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 4 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 6 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 6 \tab 1 \tab 3 \tab 75 \tab 7 \tab 22 \tab 2 \tab 22 \tab 24 \tab - \tab 1 \tab 8 \tab 5 \tab 5 \tab - \tab 9 \tab 11 \tab - \tab 1 \tab 28 \tab 4 \tab 3 \tab 2 \tab 2 \tab 26 \tab 10 \tab 3 \tab 1 \tab 3 \tab 2 \tab 31 \tab 17 \tab - \tab 1 \tab 9 \tab 2 \tab 3 \tab 29 \tab 9 \tab 16 \tab 5 \tab 2 \tab 42 \tab 21 \tab 16 \tab 9 \tab 16 \tab 11 \tab 9 \tab 6 \tab 6 \tab 1 \tab 27 \tab 6 \tab 5 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 13 \tab 7 \tab 15 \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 8 \tab 6 \tab 14 \tab 3 \tab 2 \tab 4 \tab 6 \tab 33 \tab 11 \tab 1 \tab 4 \tab - \tab 4 \tab 9 \tab 2 \tab 13 \tab 2 \tab 9 \tab 24 \tab 7 \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab 17 \tab 7 \tab 37 \tab 87 \tab 1 \tab 2 \tab 3 \tab 6 \tab 35 \tab 23 \tab - \tab 9 \tab 5 \tab 3 \tab 8 \tab 4 \tab 25 \tab 1 \tab 1 \tab - \tab 4 \tab 9 \tab 14 \tab 23 \tab 11 \tab 6 \tab - \tab 3 \tab 4 \tab 13 \tab 1 \tab 5 \tab 2 \tab 7 \tab 15 \tab 6 \tab 8 \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab 4 \tab 1 \tab 4 \tab 8 \tab 7 \tab 1 \tab 7 \tab 4 \tab 13 \tab 6 \tab 4 \tab 2 \tab 5 \tab 2 \tab 7 \tab 8 \tab 1 \tab 6 \tab 3 \tab 9 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 9 \tab - \tab 18 \tab - \tab 2 \tab 6 \tab 8 \tab 1 \tab 1 \tab 1 \tab - \tab 7 \tab 1 \tab - \tab 1 \tab - \tab 9 \tab - \tab 7 \tab 8 \tab - \tab 2 \tab 8 \tab 3 \tab 44 \tab 1 \tab 6 \tab - \tab 1 \tab 19 \tab 2 \tab 3 \tab 27 \tab 9 \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 12 \tab 1 \tab 1 \tab 2 \tab 168 \tab - \tab 3 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 3 \tab 5 \tab 3 \tab 1 \tab 4 \tab 2 \tab 4 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 8 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 8 \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 2 \tab 3 \tab 1 \tab 28 \tab 2 \tab 22 \tab 3 \tab 30 \tab 16 \tab 1 \tab 3 \tab 8 \tab 4 \tab 4 \tab 2 \tab 7 \tab 6 \tab - \tab - \tab 27 \tab 4 \tab 1 \tab 2 \tab 2 \tab 15 \tab 3 \tab 4 \tab 6 \tab 1 \tab 3 \tab 31 \tab 9 \tab 2 \tab 1 \tab 5 \tab - \tab 2 \tab 24 \tab 4 \tab 13 \tab 5 \tab 2 \tab 24 \tab 19 \tab 12 \tab 10 \tab 10 \tab 9 \tab 6 \tab - \tab 4 \tab 1 \tab 16 \tab 6 \tab 9 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 12 \tab 7 \tab 21 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 8 \tab 3 \tab - \tab 5 \tab 5 \tab 24 \tab 15 \tab 1 \tab 6 \tab 1 \tab 3 \tab 4 \tab 1 \tab 11 \tab 1 \tab 5 \tab 15 \tab 8 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab 5 \tab - \tab 18 \tab - \tab 49 \tab 102 \tab 2 \tab 1 \tab 7 \tab 1 \tab 26 \tab 32 \tab - \tab 9 \tab 4 \tab 4 \tab 5 \tab 1 \tab 23 \tab - \tab - \tab 1 \tab 1 \tab 11 \tab 4 \tab 9 \tab 9 \tab 7 \tab - \tab 4 \tab 1 \tab 6 \tab - \tab 4 \tab 2 \tab 6 \tab 23 \tab 3 \tab 4 \tab - \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 4 \tab - \tab 3 \tab 1 \tab 7 \tab - \tab 9 \tab 11 \tab 5 \tab - \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 6 \tab - \tab 5 \tab 18 \tab 2 \tab 6 \tab 2 \tab 10 \tab 6 \tab - \tab - \tab 1 \tab - \tab 7 \tab 1 \tab 10 \tab - \tab 4 \tab 7 \tab 10 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab 7 \tab 1 \tab 5 \tab 15 \tab 4 \tab 4 \tab 8 \tab 3 \tab 34 \tab 4 \tab 2 \tab - \tab 1 \tab 21 \tab 6 \tab 2 \tab 21 \tab 19 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 6 \tab - \tab 2 \tab 2 \tab 133 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 4 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 8 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 2 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 3 \tab 3 \tab 7 \tab 30 \tab - \tab 15 \tab 3 \tab 22 \tab 19 \tab - \tab 4 \tab 9 \tab 5 \tab 2 \tab 2 \tab 4 \tab 5 \tab 1 \tab 4 \tab 21 \tab 4 \tab 3 \tab - \tab 2 \tab 26 \tab 6 \tab 7 \tab - \tab 1 \tab 1 \tab 26 \tab 12 \tab 3 \tab 3 \tab 1 \tab - \tab 1 \tab 18 \tab 7 \tab 15 \tab 3 \tab 1 \tab 20 \tab 14 \tab 14 \tab 6 \tab 10 \tab 6 \tab 10 \tab 5 \tab 5 \tab - \tab 18 \tab 4 \tab 11 \tab 2 \tab - \tab 4 \tab 3 \tab 1 \tab 6 \tab 5 \tab 19 \tab - \tab - \tab - \tab 3 \tab - \tab 7 \tab 1 \tab 6 \tab 2 \tab - \tab 5 \tab 4 \tab 23 \tab 11 \tab - \tab 2 \tab 1 \tab 3 \tab 5 \tab - \tab 8 \tab 1 \tab 4 \tab 14 \tab 6 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 9 \tab - \tab 15 \tab 3 \tab 37 \tab 68 \tab 1 \tab 6 \tab 7 \tab 3 \tab 53 \tab 22 \tab 1 \tab 8 \tab 9 \tab 6 \tab 7 \tab 1 \tab 30 \tab - \tab 3 \tab 2 \tab 4 \tab 4 \tab 5 \tab 13 \tab 12 \tab 5 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 5 \tab 1 \tab 4 \tab 22 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 5 \tab - \tab 1 \tab 2 \tab 6 \tab - \tab 10 \tab 10 \tab 5 \tab 1 \tab 1 \tab 1 \tab 8 \tab 5 \tab 2 \tab 1 \tab 2 \tab 3 \tab 6 \tab 11 \tab - \tab 2 \tab 2 \tab 5 \tab 6 \tab 3 \tab - \tab 1 \tab - \tab 11 \tab - \tab 14 \tab 2 \tab 2 \tab 9 \tab 10 \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 4 \tab 3 \tab 2 \tab 13 \tab 1 \tab 3 \tab 9 \tab 1 \tab 18 \tab 7 \tab 3 \tab - \tab - \tab 25 \tab 1 \tab 1 \tab 24 \tab 8 \tab 4 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 5 \tab 5 \tab 1 \tab - \tab 5 \tab 95 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 6 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab 4 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 2 \tab 1 \tab 4 \tab 25 \tab 3 \tab 16 \tab 1 \tab 20 \tab 21 \tab - \tab 2 \tab 7 \tab 3 \tab 2 \tab 2 \tab 12 \tab 8 \tab - \tab - \tab 34 \tab 4 \tab - \tab 1 \tab 2 \tab 20 \tab 8 \tab 5 \tab - \tab - \tab 2 \tab 15 \tab 16 \tab - \tab 1 \tab 3 \tab 1 \tab 4 \tab 18 \tab 3 \tab 18 \tab 5 \tab 2 \tab 17 \tab 17 \tab 7 \tab 3 \tab 8 \tab 11 \tab 5 \tab 1 \tab 3 \tab - \tab 7 \tab 6 \tab 7 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 9 \tab 3 \tab 16 \tab 1 \tab - \tab - \tab - \tab 1 \tab 7 \tab 3 \tab 9 \tab 1 \tab - \tab 2 \tab 8 \tab 33 \tab 10 \tab 2 \tab 3 \tab 2 \tab 1 \tab 9 \tab 2 \tab 10 \tab 3 \tab 7 \tab 20 \tab 5 \tab 4 \tab 3 \tab - \tab 5 \tab 2 \tab - \tab 9 \tab 1 \tab 9 \tab 4 \tab 40 \tab 80 \tab 1 \tab 1 \tab 2 \tab 2 \tab 24 \tab 25 \tab - \tab 6 \tab 4 \tab 4 \tab 10 \tab 3 \tab 28 \tab 1 \tab 1 \tab 3 \tab 6 \tab 16 \tab 5 \tab 17 \tab 11 \tab 11 \tab - \tab 4 \tab 2 \tab 1 \tab 3 \tab 4 \tab 1 \tab 2 \tab 24 \tab 3 \tab 7 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 4 \tab 4 \tab - \tab 1 \tab 5 \tab 9 \tab - \tab 2 \tab 10 \tab 8 \tab 4 \tab 1 \tab 4 \tab 15 \tab 6 \tab 5 \tab 2 \tab 4 \tab 3 \tab 6 \tab 13 \tab 1 \tab 3 \tab 1 \tab 3 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 14 \tab - \tab 22 \tab - \tab 4 \tab 5 \tab 12 \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 2 \tab 1 \tab 7 \tab 1 \tab 8 \tab 11 \tab - \tab 1 \tab 8 \tab 2 \tab 26 \tab 4 \tab 4 \tab - \tab - \tab 19 \tab 3 \tab 3 \tab 22 \tab 9 \tab 6 \tab 4 \tab 4 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 9 \tab - \tab 6 \tab 6 \tab 120 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 5 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 4 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 2 \tab 5 \tab 23 \tab 1 \tab 13 \tab 4 \tab 18 \tab 16 \tab 2 \tab 2 \tab 9 \tab 7 \tab 5 \tab 1 \tab 7 \tab 2 \tab - \tab - \tab 23 \tab 7 \tab 1 \tab 1 \tab - \tab 22 \tab 9 \tab 3 \tab 1 \tab - \tab 1 \tab 20 \tab 11 \tab 1 \tab - \tab 5 \tab - \tab 3 \tab 22 \tab 2 \tab 9 \tab 5 \tab 2 \tab 27 \tab 23 \tab 12 \tab 4 \tab 4 \tab 12 \tab 7 \tab 5 \tab 1 \tab - \tab 27 \tab 9 \tab 9 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 7 \tab 7 \tab 22 \tab 1 \tab - \tab - \tab 5 \tab 2 \tab 2 \tab 6 \tab 7 \tab 3 \tab - \tab 6 \tab 4 \tab 22 \tab 11 \tab - \tab 4 \tab - \tab 1 \tab 10 \tab - \tab 11 \tab 2 \tab 8 \tab 14 \tab 6 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 15 \tab - \tab 31 \tab 70 \tab 1 \tab - \tab 2 \tab 1 \tab 44 \tab 22 \tab - \tab 10 \tab 3 \tab 2 \tab 3 \tab 4 \tab 30 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab 14 \tab 15 \tab 11 \tab 6 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 7 \tab 1 \tab 4 \tab 14 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 7 \tab - \tab 1 \tab - \tab 7 \tab 1 \tab 8 \tab 13 \tab 7 \tab - \tab 2 \tab 2 \tab 7 \tab 3 \tab 4 \tab 1 \tab 2 \tab 2 \tab 5 \tab 13 \tab 1 \tab 1 \tab 3 \tab 3 \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab 11 \tab - \tab 20 \tab - \tab 7 \tab 2 \tab 11 \tab - \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 4 \tab 3 \tab 6 \tab 10 \tab 1 \tab 4 \tab 6 \tab 3 \tab 27 \tab 3 \tab - \tab - \tab 1 \tab 17 \tab 5 \tab 2 \tab 30 \tab 18 \tab 3 \tab 6 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 6 \tab - \tab 2 \tab 7 \tab 112 \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 6 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 2 \tab 3 \tab 3 \tab 61 \tab 2 \tab 17 \tab 5 \tab 28 \tab 19 \tab - \tab - \tab 10 \tab 3 \tab 6 \tab 2 \tab 6 \tab 3 \tab - \tab 1 \tab 19 \tab 7 \tab 1 \tab 3 \tab 5 \tab 29 \tab 6 \tab 3 \tab 3 \tab 3 \tab - \tab 32 \tab 14 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 29 \tab 3 \tab 20 \tab 4 \tab 1 \tab 50 \tab 26 \tab 10 \tab 2 \tab 16 \tab 10 \tab 12 \tab 3 \tab 7 \tab - \tab 16 \tab 2 \tab 9 \tab 1 \tab - \tab - \tab 3 \tab 5 \tab 10 \tab 6 \tab 28 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 6 \tab 9 \tab 7 \tab 1 \tab - \tab 5 \tab 6 \tab 27 \tab 16 \tab 2 \tab 6 \tab 2 \tab 3 \tab 12 \tab 3 \tab 5 \tab 1 \tab 7 \tab 15 \tab 4 \tab 3 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 7 \tab 1 \tab 16 \tab 3 \tab 40 \tab 80 \tab 2 \tab 2 \tab 2 \tab 5 \tab 39 \tab 30 \tab - \tab 8 \tab 5 \tab - \tab 5 \tab 3 \tab 36 \tab - \tab - \tab - \tab 4 \tab 8 \tab 9 \tab 21 \tab 21 \tab 4 \tab 1 \tab 3 \tab - \tab 8 \tab 1 \tab 3 \tab 3 \tab 6 \tab 16 \tab 2 \tab 6 \tab 4 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 5 \tab - \tab 2 \tab 1 \tab 7 \tab 2 \tab 6 \tab 10 \tab 7 \tab - \tab 2 \tab 7 \tab 12 \tab 5 \tab 3 \tab - \tab 5 \tab 1 \tab 10 \tab 11 \tab 1 \tab 5 \tab 3 \tab 11 \tab 3 \tab 2 \tab 2 \tab 3 \tab - \tab 7 \tab - \tab 23 \tab 2 \tab 5 \tab 5 \tab 4 \tab 2 \tab 3 \tab 2 \tab 1 \tab 2 \tab 8 \tab 1 \tab 1 \tab 1 \tab 9 \tab 2 \tab 3 \tab 11 \tab 1 \tab 4 \tab 8 \tab 3 \tab 25 \tab 4 \tab 4 \tab 1 \tab - \tab 26 \tab 6 \tab 3 \tab 23 \tab 13 \tab 6 \tab 4 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 4 \tab 7 \tab - \tab 1 \tab 4 \tab 177 \tab 1 \tab 4 \tab 1 \tab - \tab 2 \tab 3 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1 \tab 4 \tab 3 \tab 35 \tab 1 \tab 24 \tab 4 \tab 25 \tab 9 \tab 1 \tab 3 \tab 6 \tab 2 \tab 2 \tab 2 \tab 7 \tab 8 \tab 1 \tab 1 \tab 18 \tab 2 \tab 2 \tab 1 \tab 3 \tab 20 \tab 6 \tab 4 \tab 2 \tab 2 \tab 2 \tab 19 \tab 16 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab 21 \tab 8 \tab 13 \tab 5 \tab 1 \tab 26 \tab 11 \tab 19 \tab - \tab 12 \tab 8 \tab 8 \tab 2 \tab 1 \tab 1 \tab 13 \tab 5 \tab 8 \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab 3 \tab 5 \tab 18 \tab 2 \tab - \tab 1 \tab 4 \tab 1 \tab 4 \tab 3 \tab 10 \tab - \tab - \tab 11 \tab 5 \tab 21 \tab 9 \tab - \tab 5 \tab 1 \tab 2 \tab 3 \tab 1 \tab 11 \tab - \tab 4 \tab 14 \tab 1 \tab - \tab - \tab - \tab 7 \tab - \tab 1 \tab 7 \tab - \tab 8 \tab 3 \tab 31 \tab 72 \tab - \tab - \tab 7 \tab 2 \tab 52 \tab 10 \tab - \tab 5 \tab 6 \tab 6 \tab 5 \tab 2 \tab 42 \tab 3 \tab 1 \tab - \tab 8 \tab 7 \tab 8 \tab 14 \tab 10 \tab 2 \tab - \tab 2 \tab 2 \tab 4 \tab 2 \tab 5 \tab - \tab 3 \tab 20 \tab 5 \tab 7 \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab 3 \tab - \tab 2 \tab 1 \tab 6 \tab 1 \tab 2 \tab 9 \tab 3 \tab 3 \tab 2 \tab 6 \tab 7 \tab 7 \tab 4 \tab 1 \tab 3 \tab 1 \tab 6 \tab 21 \tab 3 \tab 6 \tab 7 \tab 6 \tab 6 \tab 2 \tab 1 \tab 4 \tab - \tab 8 \tab 1 \tab 14 \tab 3 \tab 1 \tab 10 \tab 10 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 5 \tab 2 \tab 1 \tab 3 \tab 6 \tab 1 \tab 2 \tab 11 \tab - \tab 2 \tab 5 \tab 3 \tab 25 \tab 7 \tab 3 \tab - \tab 1 \tab 11 \tab 6 \tab - \tab 24 \tab 13 \tab 5 \tab 5 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 6 \tab 89 \tab 4 \tab 3 \tab - \tab - \tab 3 \tab 3 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 4 \tab 4 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab 4 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1 \tab 3 \tab 2 \tab 24 \tab - \tab 11 \tab 5 \tab 22 \tab 14 \tab 1 \tab 1 \tab 6 \tab 7 \tab 1 \tab 2 \tab 9 \tab 6 \tab - \tab 4 \tab 26 \tab 3 \tab 1 \tab 1 \tab 1 \tab 25 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 23 \tab 9 \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 19 \tab 6 \tab 14 \tab 5 \tab - \tab 29 \tab 7 \tab 9 \tab 7 \tab 7 \tab 14 \tab 5 \tab 2 \tab 6 \tab - \tab 22 \tab 1 \tab 10 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 13 \tab 2 \tab 17 \tab 2 \tab - \tab - \tab - \tab 1 \tab 4 \tab 7 \tab 8 \tab 1 \tab - \tab 7 \tab 3 \tab 32 \tab 8 \tab 1 \tab 4 \tab - \tab 3 \tab 5 \tab 1 \tab 8 \tab 2 \tab 6 \tab 21 \tab 6 \tab - \tab - \tab - \tab 5 \tab 1 \tab - \tab 8 \tab 1 \tab 15 \tab - \tab 28 \tab 84 \tab 1 \tab 3 \tab 4 \tab 2 \tab 47 \tab 20 \tab - \tab 7 \tab 9 \tab 3 \tab 6 \tab - \tab 38 \tab 2 \tab 2 \tab 3 \tab 4 \tab 10 \tab 9 \tab 10 \tab 14 \tab 3 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 7 \tab 24 \tab 3 \tab 5 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 1 \tab 7 \tab 2 \tab 5 \tab 2 \tab 5 \tab 1 \tab 5 \tab 8 \tab 7 \tab - \tab 1 \tab 2 \tab 10 \tab 4 \tab 2 \tab 1 \tab 4 \tab 3 \tab 9 \tab 15 \tab 1 \tab 3 \tab 1 \tab 7 \tab 7 \tab 1 \tab - \tab 2 \tab - \tab 7 \tab - \tab 23 \tab 1 \tab 3 \tab 12 \tab 9 \tab 1 \tab - \tab 4 \tab 2 \tab 6 \tab 6 \tab 2 \tab 2 \tab 2 \tab 9 \tab 2 \tab 7 \tab 8 \tab 3 \tab 4 \tab 8 \tab 2 \tab 19 \tab 3 \tab 3 \tab - \tab - \tab 23 \tab 9 \tab 4 \tab 24 \tab 17 \tab 4 \tab 6 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 12 \tab - \tab 1 \tab 4 \tab 109 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 5 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 3 \tab 3 \tab 2 \tab 20 \tab 2 \tab 13 \tab 4 \tab 23 \tab 18 \tab - \tab 2 \tab 4 \tab 2 \tab 5 \tab 1 \tab 6 \tab 3 \tab - \tab - \tab 19 \tab 2 \tab 3 \tab 1 \tab - \tab 11 \tab 3 \tab 6 \tab 1 \tab - \tab - \tab 16 \tab 12 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 16 \tab 9 \tab 7 \tab 3 \tab - \tab 23 \tab 16 \tab 8 \tab 3 \tab 13 \tab 13 \tab 12 \tab - \tab 6 \tab - \tab 19 \tab 5 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 5 \tab 4 \tab 11 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 6 \tab 1 \tab - \tab 9 \tab 3 \tab 29 \tab 7 \tab 2 \tab 2 \tab 1 \tab 3 \tab 3 \tab 1 \tab 4 \tab - \tab 7 \tab 17 \tab 10 \tab 6 \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab 3 \tab 1 \tab 12 \tab 1 \tab 31 \tab 77 \tab 2 \tab 1 \tab 5 \tab 1 \tab 28 \tab 19 \tab - \tab 13 \tab - \tab 2 \tab 7 \tab 2 \tab 31 \tab 3 \tab 1 \tab 1 \tab 5 \tab 5 \tab 15 \tab 8 \tab 10 \tab 2 \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 3 \tab - \tab 5 \tab 19 \tab 3 \tab 4 \tab 2 \tab - \tab - \tab - \tab 2 \tab 4 \tab 4 \tab - \tab 1 \tab - \tab 13 \tab - \tab 9 \tab 11 \tab 7 \tab - \tab 1 \tab 3 \tab 14 \tab 7 \tab 4 \tab - \tab 3 \tab 3 \tab 10 \tab 16 \tab 3 \tab 3 \tab 1 \tab 7 \tab 7 \tab 1 \tab 1 \tab 2 \tab 1 \tab 9 \tab - \tab 16 \tab 1 \tab 3 \tab 11 \tab 7 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 7 \tab 2 \tab 8 \tab 13 \tab - \tab 3 \tab 4 \tab 2 \tab 23 \tab 2 \tab 5 \tab - \tab 3 \tab 20 \tab 3 \tab 4 \tab 20 \tab 13 \tab 2 \tab 10 \tab - \tab 1 \tab 4 \tab - \tab - \tab 2 \tab 5 \tab - \tab 1 \tab 4 \tab 73 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab 2 \tab 5 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1 \tab - \tab 5 \tab 12 \tab 3 \tab 27 \tab 4 \tab 25 \tab 16 \tab - \tab 4 \tab 2 \tab 3 \tab 8 \tab 2 \tab 7 \tab 4 \tab - \tab 1 \tab 25 \tab 3 \tab - \tab 1 \tab 4 \tab 21 \tab 7 \tab - \tab - \tab 1 \tab 1 \tab 24 \tab 12 \tab - \tab - \tab 3 \tab 2 \tab - \tab 19 \tab 3 \tab 15 \tab 1 \tab 3 \tab 38 \tab 13 \tab 8 \tab 8 \tab 8 \tab 13 \tab 7 \tab 5 \tab 3 \tab 2 \tab 25 \tab 3 \tab 9 \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab 8 \tab 4 \tab 18 \tab 2 \tab - \tab - \tab 2 \tab 4 \tab 6 \tab 7 \tab 7 \tab 2 \tab - \tab 5 \tab 5 \tab 24 \tab 18 \tab 1 \tab 4 \tab 2 \tab 1 \tab 8 \tab 2 \tab 9 \tab - \tab 4 \tab 19 \tab 6 \tab 1 \tab 1 \tab - \tab 6 \tab - \tab - \tab 8 \tab 1 \tab 15 \tab 5 \tab 36 \tab 68 \tab - \tab 4 \tab 3 \tab 1 \tab 36 \tab 14 \tab - \tab 4 \tab 10 \tab 2 \tab 2 \tab 1 \tab 34 \tab - \tab 1 \tab 2 \tab 4 \tab 6 \tab 7 \tab 13 \tab 13 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 6 \tab 1 \tab 1 \tab 16 \tab 6 \tab 7 \tab 5 \tab - \tab - \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 5 \tab 3 \tab 8 \tab - \tab 3 \tab 8 \tab 5 \tab - \tab 1 \tab 4 \tab 10 \tab 5 \tab 3 \tab 1 \tab 4 \tab - \tab 12 \tab 12 \tab - \tab 2 \tab 2 \tab 7 \tab 6 \tab 1 \tab - \tab 2 \tab - \tab 9 \tab - \tab 25 \tab 2 \tab 3 \tab 7 \tab 12 \tab 2 \tab - \tab 2 \tab 2 \tab 5 \tab 2 \tab 3 \tab 7 \tab - \tab 9 \tab 1 \tab 4 \tab 9 \tab 1 \tab - \tab 6 \tab 1 \tab 30 \tab 3 \tab 2 \tab - \tab - \tab 20 \tab 5 \tab 4 \tab 18 \tab 12 \tab 2 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab 6 \tab 1 \tab - \tab 4 \tab 100 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 7 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 3 \tab 2 \tab 3 \tab 3 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 4 \tab 2 \tab 7 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 2 \tab 4 \tab - \tab 25 \tab - \tab 20 \tab 2 \tab 21 \tab 11 \tab 2 \tab 1 \tab 5 \tab 3 \tab 3 \tab 1 \tab 9 \tab 4 \tab - \tab 2 \tab 24 \tab 1 \tab - \tab - \tab 5 \tab 20 \tab 8 \tab 7 \tab 1 \tab 2 \tab - \tab 25 \tab 12 \tab 3 \tab 1 \tab 4 \tab - \tab 2 \tab 20 \tab 9 \tab 11 \tab 3 \tab - \tab 25 \tab 17 \tab 12 \tab 4 \tab 10 \tab 8 \tab 9 \tab 6 \tab 5 \tab - \tab 23 \tab 5 \tab 7 \tab - \tab 2 \tab 2 \tab 3 \tab 1 \tab 6 \tab 7 \tab 18 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab 6 \tab 10 \tab 1 \tab - \tab 7 \tab 4 \tab 31 \tab 12 \tab 3 \tab 4 \tab - \tab 3 \tab 4 \tab 2 \tab 10 \tab 2 \tab 5 \tab 13 \tab 8 \tab 4 \tab 4 \tab - \tab 4 \tab 1 \tab - \tab 8 \tab - \tab 18 \tab 1 \tab 33 \tab 75 \tab 1 \tab 3 \tab 6 \tab 4 \tab 31 \tab 24 \tab 3 \tab 5 \tab 3 \tab 3 \tab 6 \tab 3 \tab 37 \tab 2 \tab 4 \tab 3 \tab 6 \tab 12 \tab 8 \tab 12 \tab 10 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 7 \tab 2 \tab 2 \tab 25 \tab 5 \tab 7 \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab 9 \tab 1 \tab 2 \tab 2 \tab 12 \tab 1 \tab 10 \tab 9 \tab 8 \tab - \tab - \tab 2 \tab 15 \tab 4 \tab 6 \tab - \tab 8 \tab - \tab 8 \tab 13 \tab 4 \tab 4 \tab 3 \tab 9 \tab 7 \tab 1 \tab - \tab 2 \tab - \tab 12 \tab 2 \tab 20 \tab 7 \tab 2 \tab 9 \tab 6 \tab - \tab 2 \tab 5 \tab 3 \tab 5 \tab 7 \tab 2 \tab 1 \tab 4 \tab 14 \tab 1 \tab 11 \tab 16 \tab - \tab 1 \tab 10 \tab 7 \tab 25 \tab 2 \tab 3 \tab - \tab 1 \tab 15 \tab 3 \tab 2 \tab 23 \tab 7 \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 15 \tab - \tab 1 \tab 7 \tab 85 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 4 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab 6 \tab 1 \tab 1 \tab 4 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab 4 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 3 \tab 3 \tab 5 \tab 5 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 5 \tab 4 \tab 5 \tab 48 \tab 2 \tab 17 \tab 4 \tab 27 \tab 24 \tab 7 \tab 2 \tab 8 \tab 4 \tab 7 \tab 6 \tab 13 \tab 6 \tab - \tab 2 \tab 52 \tab 14 \tab 2 \tab 2 \tab 5 \tab 37 \tab 6 \tab 9 \tab 3 \tab 3 \tab 3 \tab 47 \tab 33 \tab 4 \tab 3 \tab 5 \tab 3 \tab 1 \tab 43 \tab 16 \tab 24 \tab 7 \tab 2 \tab 59 \tab 27 \tab 18 \tab 7 \tab 18 \tab 21 \tab 24 \tab 2 \tab 7 \tab 1 \tab 46 \tab 10 \tab 10 \tab 1 \tab 2 \tab 2 \tab 2 \tab 8 \tab 15 \tab 6 \tab 34 \tab 1 \tab 3 \tab - \tab 3 \tab 2 \tab 11 \tab 11 \tab 17 \tab 3 \tab - \tab 16 \tab 8 \tab 59 \tab 21 \tab 2 \tab 8 \tab - \tab 1 \tab 10 \tab 1 \tab 11 \tab 1 \tab 12 \tab 18 \tab 17 \tab 5 \tab 7 \tab 2 \tab 13 \tab - \tab - \tab 30 \tab 1 \tab 24 \tab 2 \tab 50 \tab 149 \tab 1 \tab 3 \tab 9 \tab 10 \tab 74 \tab 25 \tab - \tab 18 \tab 8 \tab 7 \tab 10 \tab 6 \tab 60 \tab 2 \tab 2 \tab 3 \tab 18 \tab 20 \tab 16 \tab 24 \tab 20 \tab 8 \tab - \tab 3 \tab 3 \tab 3 \tab - \tab 8 \tab 2 \tab 6 \tab 31 \tab 4 \tab 9 \tab 3 \tab 1 \tab - \tab - \tab 4 \tab 5 \tab 11 \tab - \tab 5 \tab 5 \tab 15 \tab 1 \tab 14 \tab 12 \tab 8 \tab 2 \tab 6 \tab 9 \tab 23 \tab 10 \tab 10 \tab 1 \tab 8 \tab 7 \tab 11 \tab 27 \tab 3 \tab 10 \tab 5 \tab 20 \tab 13 \tab 3 \tab 3 \tab 6 \tab 1 \tab 23 \tab - \tab 46 \tab 3 \tab 8 \tab 18 \tab 18 \tab 5 \tab 3 \tab 3 \tab 4 \tab 3 \tab 12 \tab 4 \tab 4 \tab 6 \tab 17 \tab 2 \tab 10 \tab 33 \tab 3 \tab 3 \tab 26 \tab 9 \tab 61 \tab 10 \tab 8 \tab 2 \tab 2 \tab 33 \tab 11 \tab 6 \tab 39 \tab 25 \tab 8 \tab 6 \tab 4 \tab 11 \tab 5 \tab - \tab 2 \tab 4 \tab 17 \tab 1 \tab 1 \tab 9 \tab 147 \tab 1 \tab 9 \tab 2 \tab - \tab - \tab 5 \tab 2 \tab 1 \tab 6 \tab 4 \tab 2 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 10 \tab 5 \tab - \tab 6 \tab 1 \tab - \tab 5 \tab 7 \tab 1 \tab 2 \tab 3 \tab 3 \tab 4 \tab 1 \tab 2 \tab 6 \tab - \tab 4 \tab 2 \tab 4 \tab 6 \tab 11 \tab 2 \tab - \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 6 \tab 4 \tab 3 \tab - \tab 4 \tab - \tab 4 \tab 10 \tab 10 \tab 11 \tab 3 \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2992 \cr
   1996 \tab 4 \tab 7 \tab 5 \tab 43 \tab 3 \tab 24 \tab 3 \tab 46 \tab 22 \tab 3 \tab 3 \tab 8 \tab 8 \tab 6 \tab 8 \tab 21 \tab 6 \tab - \tab 2 \tab 40 \tab 4 \tab 2 \tab 6 \tab 2 \tab 49 \tab 19 \tab 11 \tab 3 \tab - \tab 3 \tab 38 \tab 30 \tab 1 \tab 3 \tab 7 \tab 11 \tab 1 \tab 42 \tab 12 \tab 37 \tab 8 \tab 3 \tab 58 \tab 26 \tab 28 \tab 11 \tab 13 \tab 37 \tab 22 \tab 5 \tab 4 \tab - \tab 37 \tab 7 \tab 17 \tab 2 \tab - \tab 3 \tab 4 \tab 5 \tab 15 \tab 4 \tab 33 \tab 3 \tab 3 \tab - \tab 1 \tab 7 \tab 4 \tab 7 \tab 21 \tab 1 \tab - \tab 15 \tab 6 \tab 55 \tab 13 \tab 1 \tab 10 \tab - \tab 4 \tab 14 \tab 6 \tab 8 \tab 4 \tab 9 \tab 35 \tab 21 \tab 5 \tab 4 \tab - \tab 8 \tab 1 \tab 1 \tab 26 \tab - \tab 22 \tab 3 \tab 59 \tab 108 \tab - \tab 1 \tab 7 \tab 1 \tab 63 \tab 16 \tab 2 \tab 22 \tab 17 \tab 7 \tab 6 \tab 1 \tab 62 \tab 2 \tab 4 \tab 4 \tab 9 \tab 9 \tab 18 \tab 37 \tab 23 \tab 4 \tab - \tab 3 \tab 2 \tab 1 \tab 4 \tab 8 \tab - \tab - \tab 35 \tab 2 \tab 11 \tab 5 \tab 3 \tab 3 \tab - \tab 6 \tab 3 \tab 11 \tab 3 \tab 3 \tab 2 \tab 11 \tab 1 \tab 18 \tab 15 \tab 7 \tab 1 \tab 3 \tab 9 \tab 27 \tab 11 \tab 7 \tab 1 \tab 5 \tab 6 \tab 20 \tab 19 \tab 3 \tab 13 \tab 4 \tab 21 \tab 11 \tab 5 \tab - \tab 4 \tab 1 \tab 24 \tab - \tab 40 \tab 3 \tab 10 \tab 23 \tab 15 \tab 3 \tab 1 \tab 4 \tab 6 \tab 9 \tab 13 \tab 2 \tab 2 \tab 2 \tab 18 \tab 3 \tab 8 \tab 26 \tab 1 \tab 6 \tab 20 \tab 7 \tab 50 \tab 9 \tab 6 \tab - \tab 1 \tab 36 \tab 11 \tab 4 \tab 40 \tab 15 \tab 9 \tab 11 \tab - \tab 11 \tab 5 \tab 4 \tab - \tab 1 \tab 10 \tab - \tab 4 \tab 16 \tab 123 \tab 5 \tab 11 \tab 2 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 3 \tab 4 \tab 2 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 6 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 12 \tab 3 \tab 1 \tab 5 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 3 \tab 8 \tab 1 \tab 2 \tab 2 \tab 12 \tab 3 \tab - \tab 3 \tab 3 \tab 2 \tab 4 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 4 \tab 6 \tab 3 \tab 1 \tab - \tab 3 \tab 3 \tab 4 \tab 13 \tab 8 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 3 \tab - \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 6 \tab 1 \tab 3 \tab - \tab 3 \tab 4 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 4 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 2 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 1 \tab 3 \tab 38 \tab 1 \tab 8 \tab 5 \tab 49 \tab 19 \tab 1 \tab 6 \tab 9 \tab 10 \tab 10 \tab 4 \tab 20 \tab 5 \tab - \tab 2 \tab 42 \tab 7 \tab 2 \tab 2 \tab 3 \tab 30 \tab 10 \tab 12 \tab 1 \tab 3 \tab 3 \tab 53 \tab 32 \tab 3 \tab 3 \tab 3 \tab 3 \tab - \tab 33 \tab 10 \tab 27 \tab 10 \tab 2 \tab 56 \tab 23 \tab 32 \tab 7 \tab 17 \tab 23 \tab 22 \tab 1 \tab 6 \tab 1 \tab 32 \tab 5 \tab 16 \tab - \tab 1 \tab 3 \tab 2 \tab 5 \tab 16 \tab 9 \tab 33 \tab - \tab 2 \tab - \tab 2 \tab 8 \tab 8 \tab 6 \tab 18 \tab - \tab - \tab 9 \tab 6 \tab 59 \tab 19 \tab - \tab 6 \tab 1 \tab 2 \tab 11 \tab 4 \tab 7 \tab 2 \tab 10 \tab 30 \tab 10 \tab - \tab 4 \tab - \tab 10 \tab 1 \tab - \tab 28 \tab 1 \tab 22 \tab 5 \tab 37 \tab 126 \tab 3 \tab 3 \tab 10 \tab 3 \tab 62 \tab 21 \tab - \tab 20 \tab 6 \tab 7 \tab 3 \tab 2 \tab 52 \tab 4 \tab 3 \tab 7 \tab 9 \tab 18 \tab 19 \tab 28 \tab 20 \tab 1 \tab - \tab 4 \tab 6 \tab 6 \tab 1 \tab 6 \tab 3 \tab 3 \tab 36 \tab 3 \tab 13 \tab 11 \tab 3 \tab 1 \tab 1 \tab 5 \tab 3 \tab 13 \tab 1 \tab 8 \tab 6 \tab 23 \tab 1 \tab 11 \tab 15 \tab 9 \tab 1 \tab 6 \tab 5 \tab 28 \tab 9 \tab 6 \tab 2 \tab 2 \tab 8 \tab 16 \tab 29 \tab 1 \tab 11 \tab 2 \tab 11 \tab 15 \tab 3 \tab 4 \tab 3 \tab - \tab 19 \tab - \tab 39 \tab 6 \tab 5 \tab 14 \tab 25 \tab 2 \tab 1 \tab 7 \tab 1 \tab 10 \tab 5 \tab 1 \tab 3 \tab 3 \tab 23 \tab 5 \tab 15 \tab 17 \tab 3 \tab 6 \tab 14 \tab 4 \tab 38 \tab 24 \tab 3 \tab 1 \tab 2 \tab 27 \tab 11 \tab 5 \tab 57 \tab 29 \tab 3 \tab 13 \tab 4 \tab 13 \tab 6 \tab 2 \tab 1 \tab 1 \tab 19 \tab - \tab 2 \tab 8 \tab 166 \tab 3 \tab 11 \tab 3 \tab - \tab 1 \tab 3 \tab 4 \tab - \tab 4 \tab 4 \tab - \tab 2 \tab 4 \tab 2 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 7 \tab 2 \tab - \tab 2 \tab 1 \tab 5 \tab 1 \tab - \tab 2 \tab 2 \tab 5 \tab 1 \tab - \tab 1 \tab 14 \tab 5 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 2 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 8 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 3 \tab 6 \tab 1 \tab 1 \tab 3 \tab 6 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 4 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 8 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 4 \tab 5 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 3 \tab 1 \tab 2 \tab 50 \tab 2 \tab 18 \tab 3 \tab 33 \tab 22 \tab - \tab 3 \tab 10 \tab 9 \tab 6 \tab 8 \tab 13 \tab 4 \tab - \tab 2 \tab 45 \tab 3 \tab 1 \tab 1 \tab 5 \tab 49 \tab 7 \tab 7 \tab 2 \tab 2 \tab 1 \tab 55 \tab 30 \tab 2 \tab 7 \tab 6 \tab 3 \tab 1 \tab 47 \tab 12 \tab 19 \tab 8 \tab 1 \tab 60 \tab 23 \tab 22 \tab 9 \tab 21 \tab 22 \tab 17 \tab 3 \tab 8 \tab 1 \tab 31 \tab 12 \tab 21 \tab 2 \tab 2 \tab - \tab 2 \tab 4 \tab 17 \tab 9 \tab 24 \tab 1 \tab 2 \tab - \tab 5 \tab 5 \tab 11 \tab 9 \tab 16 \tab 4 \tab - \tab 16 \tab 2 \tab 47 \tab 18 \tab 2 \tab 10 \tab 1 \tab 1 \tab 11 \tab 4 \tab 9 \tab 1 \tab 5 \tab 17 \tab 17 \tab 7 \tab 5 \tab 1 \tab 11 \tab - \tab - \tab 28 \tab 2 \tab 25 \tab 5 \tab 44 \tab 109 \tab 2 \tab 1 \tab 14 \tab 7 \tab 63 \tab 19 \tab 2 \tab 16 \tab 12 \tab 5 \tab 8 \tab 4 \tab 59 \tab 1 \tab 7 \tab 3 \tab 13 \tab 15 \tab 9 \tab 19 \tab 22 \tab 3 \tab - \tab - \tab 4 \tab 4 \tab 3 \tab 10 \tab 3 \tab 4 \tab 41 \tab 1 \tab 10 \tab 6 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 13 \tab - \tab 7 \tab 2 \tab 19 \tab - \tab 11 \tab 15 \tab 6 \tab 4 \tab 1 \tab 7 \tab 21 \tab 14 \tab 5 \tab 1 \tab 6 \tab 5 \tab 13 \tab 13 \tab 3 \tab 9 \tab 2 \tab 13 \tab 24 \tab 2 \tab 3 \tab 2 \tab 1 \tab 17 \tab 2 \tab 38 \tab 8 \tab 8 \tab 15 \tab 25 \tab - \tab 1 \tab 7 \tab 4 \tab 11 \tab 9 \tab 3 \tab 3 \tab 1 \tab 23 \tab 4 \tab 11 \tab 36 \tab 2 \tab 6 \tab 15 \tab 16 \tab 56 \tab 10 \tab 3 \tab - \tab - \tab 33 \tab 10 \tab 5 \tab 48 \tab 29 \tab 5 \tab 8 \tab 2 \tab 9 \tab 2 \tab 1 \tab 2 \tab 4 \tab 12 \tab - \tab 3 \tab 5 \tab 143 \tab 3 \tab 15 \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab - \tab 4 \tab 2 \tab 2 \tab 4 \tab 4 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 4 \tab 4 \tab 3 \tab 3 \tab 2 \tab - \tab - \tab 8 \tab 6 \tab 5 \tab - \tab 4 \tab 1 \tab 4 \tab 2 \tab 3 \tab 3 \tab 2 \tab 6 \tab - \tab 1 \tab 3 \tab 8 \tab 7 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab 2 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 7 \tab 9 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 3 \tab 5 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 5 \tab 1 \tab 2 \tab - \tab 1 \tab 5 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 6 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 7 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2817 \cr
   2002 \tab 2 \tab 2 \tab 4 \tab 33 \tab 2 \tab 12 \tab 6 \tab 41 \tab 29 \tab 4 \tab 1 \tab 8 \tab 4 \tab 7 \tab 4 \tab 14 \tab 2 \tab 1 \tab 2 \tab 43 \tab 1 \tab 7 \tab - \tab 6 \tab 47 \tab 6 \tab 8 \tab 3 \tab 4 \tab - \tab 53 \tab 40 \tab 2 \tab - \tab 5 \tab - \tab 2 \tab 32 \tab 11 \tab 21 \tab 12 \tab 4 \tab 60 \tab 18 \tab 23 \tab 5 \tab 20 \tab 26 \tab 20 \tab 2 \tab 6 \tab - \tab 25 \tab 6 \tab 12 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 9 \tab 12 \tab 32 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 7 \tab 12 \tab 15 \tab 1 \tab 1 \tab 14 \tab 5 \tab 49 \tab 27 \tab 2 \tab 7 \tab 1 \tab 5 \tab 15 \tab 3 \tab 9 \tab 1 \tab 6 \tab 35 \tab 11 \tab 5 \tab 2 \tab 3 \tab 12 \tab 1 \tab 1 \tab 38 \tab - \tab 20 \tab 6 \tab 37 \tab 110 \tab 4 \tab 4 \tab 6 \tab 5 \tab 72 \tab 12 \tab 4 \tab 16 \tab 9 \tab 4 \tab 9 \tab 9 \tab 48 \tab 4 \tab 4 \tab 3 \tab 8 \tab 11 \tab 18 \tab 24 \tab 9 \tab 4 \tab 1 \tab 5 \tab 5 \tab 2 \tab - \tab 6 \tab 5 \tab 4 \tab 35 \tab 8 \tab 19 \tab 6 \tab - \tab 1 \tab 1 \tab 2 \tab 8 \tab 11 \tab - \tab 7 \tab 1 \tab 19 \tab 1 \tab 8 \tab 21 \tab 9 \tab - \tab 4 \tab 4 \tab 22 \tab 14 \tab 9 \tab - \tab 5 \tab 6 \tab 14 \tab 33 \tab 1 \tab 4 \tab 5 \tab 19 \tab 22 \tab 6 \tab 1 \tab 9 \tab 1 \tab 22 \tab 1 \tab 46 \tab 3 \tab 4 \tab 15 \tab 24 \tab - \tab 2 \tab 5 \tab 2 \tab 4 \tab 10 \tab 7 \tab 6 \tab 4 \tab 19 \tab 4 \tab 16 \tab 25 \tab 1 \tab 5 \tab 15 \tab 7 \tab 50 \tab 8 \tab 6 \tab 1 \tab 1 \tab 35 \tab 12 \tab 3 \tab 34 \tab 25 \tab 6 \tab 7 \tab 1 \tab 12 \tab 3 \tab - \tab - \tab - \tab 12 \tab - \tab 1 \tab 9 \tab 112 \tab - \tab 16 \tab - \tab - \tab 2 \tab 7 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 6 \tab 4 \tab 2 \tab 3 \tab - \tab 2 \tab 5 \tab 3 \tab 3 \tab 5 \tab 1 \tab 12 \tab 4 \tab - \tab - \tab 16 \tab 4 \tab - \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 4 \tab 5 \tab 4 \tab - \tab 1 \tab 8 \tab 2 \tab 12 \tab 7 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 3 \tab 4 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 5 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab - \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 6 \tab - \tab 3 \tab 3 \tab 7 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 7 \tab 2 \tab 1 \tab 31 \tab 1 \tab 15 \tab 6 \tab 35 \tab 23 \tab 1 \tab 3 \tab 11 \tab 3 \tab 6 \tab 1 \tab 12 \tab 5 \tab - \tab 4 \tab 48 \tab 5 \tab 5 \tab 2 \tab 2 \tab 39 \tab 5 \tab 11 \tab 2 \tab 2 \tab - \tab 40 \tab 20 \tab 2 \tab 2 \tab 8 \tab 3 \tab 2 \tab 29 \tab 5 \tab 12 \tab 6 \tab 6 \tab 73 \tab 18 \tab 13 \tab 9 \tab 23 \tab 14 \tab 24 \tab 4 \tab 8 \tab 1 \tab 31 \tab 9 \tab 17 \tab 3 \tab 2 \tab 2 \tab 2 \tab 7 \tab 13 \tab 7 \tab 36 \tab 2 \tab - \tab - \tab 2 \tab 5 \tab 9 \tab 6 \tab 16 \tab 1 \tab - \tab 17 \tab 2 \tab 56 \tab 14 \tab - \tab 16 \tab 2 \tab 3 \tab 9 \tab 2 \tab 3 \tab - \tab 10 \tab 17 \tab 11 \tab 5 \tab 4 \tab - \tab 12 \tab 2 \tab 1 \tab 30 \tab 1 \tab 16 \tab 8 \tab 49 \tab 109 \tab 2 \tab 3 \tab 8 \tab 2 \tab 60 \tab 26 \tab 2 \tab 14 \tab 15 \tab 3 \tab 6 \tab 7 \tab 57 \tab 2 \tab 3 \tab 2 \tab 12 \tab 9 \tab 18 \tab 14 \tab 23 \tab 4 \tab - \tab 1 \tab 2 \tab 9 \tab 1 \tab 6 \tab 6 \tab 3 \tab 30 \tab 5 \tab 11 \tab 4 \tab 1 \tab - \tab 1 \tab 4 \tab 4 \tab 9 \tab 1 \tab 5 \tab 5 \tab 13 \tab - \tab 11 \tab 13 \tab 12 \tab - \tab 4 \tab 3 \tab 37 \tab 5 \tab 5 \tab 3 \tab 9 \tab 4 \tab 16 \tab 30 \tab 4 \tab 12 \tab 4 \tab 19 \tab 24 \tab 4 \tab 2 \tab 8 \tab 1 \tab 17 \tab 1 \tab 58 \tab 8 \tab 9 \tab 17 \tab 19 \tab 2 \tab 2 \tab 6 \tab 1 \tab 9 \tab 8 \tab 4 \tab 3 \tab 3 \tab 14 \tab 2 \tab 20 \tab 33 \tab 4 \tab 5 \tab 17 \tab 8 \tab 59 \tab 7 \tab 5 \tab 1 \tab 1 \tab 37 \tab 13 \tab 3 \tab 29 \tab 22 \tab 6 \tab 8 \tab - \tab 19 \tab 5 \tab 2 \tab 1 \tab - \tab 16 \tab - \tab 2 \tab 13 \tab 154 \tab 3 \tab 8 \tab 4 \tab - \tab 2 \tab 5 \tab 2 \tab 3 \tab 4 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab 3 \tab 7 \tab 5 \tab - \tab 2 \tab 2 \tab 13 \tab 2 \tab 1 \tab 4 \tab 3 \tab 14 \tab 1 \tab 3 \tab 2 \tab 13 \tab 9 \tab 1 \tab 2 \tab 2 \tab 1 \tab 6 \tab 3 \tab 3 \tab 5 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab 8 \tab - \tab 8 \tab 8 \tab 2 \tab 2 \tab 4 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 5 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 6 \tab - \tab 2 \tab - \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 5 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 4 \tab 3 \tab 2 \tab - \tab - \tab - \tab 5 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 2812 \cr
   2006 \tab 13 \tab 5 \tab 3 \tab 74 \tab - \tab 20 \tab 9 \tab 73 \tab 38 \tab 4 \tab 5 \tab 14 \tab 4 \tab 9 \tab 8 \tab 29 \tab 8 \tab - \tab 6 \tab 63 \tab 7 \tab 4 \tab 7 \tab 12 \tab 71 \tab 26 \tab 15 \tab - \tab 4 \tab 2 \tab 73 \tab 48 \tab - \tab 1 \tab 6 \tab 6 \tab 4 \tab 51 \tab 13 \tab 40 \tab 12 \tab 9 \tab 93 \tab 38 \tab 33 \tab 7 \tab 56 \tab 25 \tab 36 \tab 4 \tab 11 \tab 2 \tab 45 \tab 13 \tab 29 \tab 7 \tab - \tab 1 \tab 2 \tab 4 \tab 18 \tab 16 \tab 49 \tab 1 \tab 5 \tab - \tab 3 \tab 4 \tab 16 \tab 9 \tab 29 \tab 4 \tab 1 \tab 17 \tab 8 \tab 100 \tab 30 \tab 2 \tab 10 \tab 1 \tab 5 \tab 19 \tab 7 \tab 16 \tab 1 \tab 18 \tab 34 \tab 20 \tab 5 \tab 4 \tab 1 \tab 19 \tab - \tab 3 \tab 58 \tab 2 \tab 28 \tab 11 \tab 76 \tab 155 \tab 1 \tab 5 \tab 16 \tab 5 \tab 85 \tab 35 \tab - \tab 14 \tab 21 \tab 13 \tab 13 \tab 13 \tab 81 \tab 4 \tab 4 \tab 2 \tab 16 \tab 20 \tab 27 \tab 33 \tab 28 \tab 5 \tab - \tab 3 \tab 3 \tab 13 \tab 6 \tab 12 \tab 1 \tab 12 \tab 52 \tab 8 \tab 15 \tab 3 \tab - \tab 2 \tab - \tab 3 \tab 5 \tab 14 \tab 1 \tab 11 \tab 3 \tab 34 \tab 1 \tab 13 \tab 15 \tab 23 \tab 2 \tab 5 \tab 8 \tab 36 \tab 15 \tab 13 \tab 1 \tab 10 \tab 7 \tab 33 \tab 43 \tab 1 \tab 12 \tab 5 \tab 31 \tab 33 \tab 8 \tab 1 \tab 9 \tab 1 \tab 38 \tab 3 \tab 77 \tab 9 \tab 8 \tab 34 \tab 33 \tab 3 \tab 1 \tab 10 \tab 4 \tab 13 \tab 21 \tab 7 \tab 5 \tab 8 \tab 28 \tab 8 \tab 17 \tab 40 \tab 2 \tab 9 \tab 19 \tab 13 \tab 114 \tab 23 \tab 10 \tab 4 \tab 1 \tab 42 \tab 17 \tab 7 \tab 49 \tab 47 \tab 10 \tab 11 \tab 5 \tab 30 \tab 8 \tab 10 \tab - \tab - \tab 29 \tab 1 \tab 2 \tab 18 \tab 223 \tab 2 \tab 20 \tab 3 \tab 1 \tab 2 \tab 3 \tab 4 \tab 2 \tab 4 \tab 4 \tab - \tab 2 \tab 5 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 9 \tab 7 \tab 1 \tab 3 \tab - \tab 3 \tab 12 \tab 1 \tab 6 \tab 1 \tab 11 \tab 3 \tab 16 \tab 1 \tab 3 \tab 8 \tab - \tab 15 \tab 2 \tab 5 \tab - \tab 13 \tab 7 \tab - \tab 3 \tab 2 \tab 8 \tab 3 \tab 2 \tab 6 \tab 5 \tab 1 \tab 1 \tab 1 \tab - \tab 10 \tab 10 \tab 2 \tab 2 \tab 2 \tab 1 \tab 2 \tab 9 \tab 7 \tab 7 \tab 2 \tab 2 \tab 5 \tab 5 \tab 1 \tab 6 \tab - \tab 5 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 13 \tab 1 \tab 4 \tab 2 \tab 4 \tab 2 \tab 7 \tab - \tab 1 \tab 1 \tab 4 \tab 4 \tab 1 \tab 3 \tab 6 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 6 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 4 \tab 1 \tab 1 \tab 5 \tab - \tab 6 \tab 6 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 4 \tab 5 \tab 2 \tab 2 \tab 3 \tab - \tab - \tab - \tab 2 \tab 4 \tab 6 \tab 6 \tab 1 \tab - \tab 2 \tab 6 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 4510 \cr
   2008 \tab 4 \tab 3 \tab 1 \tab 22 \tab 1 \tab 14 \tab 7 \tab 35 \tab 29 \tab 3 \tab 1 \tab 2 \tab 4 \tab 2 \tab 3 \tab 9 \tab 9 \tab - \tab - \tab 33 \tab 6 \tab 4 \tab - \tab 6 \tab 25 \tab 4 \tab 12 \tab 2 \tab 4 \tab 3 \tab 32 \tab 21 \tab - \tab - \tab 1 \tab 3 \tab - \tab 21 \tab 10 \tab 15 \tab 5 \tab 4 \tab 48 \tab 15 \tab 8 \tab 6 \tab 18 \tab 17 \tab 7 \tab 2 \tab 1 \tab 1 \tab 21 \tab 8 \tab 9 \tab 1 \tab 4 \tab 1 \tab 1 \tab 3 \tab 13 \tab 8 \tab 38 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 2 \tab 3 \tab 13 \tab 4 \tab - \tab 10 \tab 4 \tab 47 \tab 11 \tab 4 \tab 8 \tab - \tab 1 \tab 13 \tab 2 \tab 5 \tab 3 \tab 7 \tab 11 \tab 4 \tab 6 \tab 2 \tab 1 \tab 10 \tab 1 \tab - \tab 29 \tab 2 \tab 18 \tab 6 \tab 16 \tab 65 \tab 1 \tab 3 \tab 6 \tab 1 \tab 32 \tab 14 \tab - \tab 10 \tab 12 \tab 5 \tab 5 \tab 2 \tab 30 \tab 1 \tab 3 \tab 1 \tab 10 \tab 6 \tab 15 \tab 16 \tab 14 \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 3 \tab 4 \tab 22 \tab 2 \tab 10 \tab 3 \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab 9 \tab - \tab 3 \tab 6 \tab 10 \tab - \tab 6 \tab 13 \tab 8 \tab - \tab - \tab 6 \tab 26 \tab 5 \tab 4 \tab - \tab 3 \tab 2 \tab 10 \tab 23 \tab 1 \tab 9 \tab 6 \tab 17 \tab 17 \tab 5 \tab 3 \tab 3 \tab - \tab 14 \tab 1 \tab 34 \tab 2 \tab 3 \tab 6 \tab 21 \tab 1 \tab 1 \tab 3 \tab - \tab 4 \tab 4 \tab 3 \tab 2 \tab 3 \tab 22 \tab 2 \tab 12 \tab 22 \tab - \tab 5 \tab 15 \tab 7 \tab 31 \tab 7 \tab 3 \tab 2 \tab 1 \tab 24 \tab 12 \tab 3 \tab 31 \tab 23 \tab 7 \tab 5 \tab 1 \tab 13 \tab 4 \tab 7 \tab 3 \tab - \tab 10 \tab - \tab 2 \tab 3 \tab 72 \tab 2 \tab 12 \tab - \tab 1 \tab 1 \tab 6 \tab 3 \tab 2 \tab - \tab 2 \tab - \tab 4 \tab 3 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 8 \tab 4 \tab - \tab - \tab - \tab 1 \tab 3 \tab 5 \tab 2 \tab 2 \tab 2 \tab 1 \tab 7 \tab 1 \tab - \tab 8 \tab 1 \tab 5 \tab - \tab 2 \tab 2 \tab 14 \tab 4 \tab - \tab - \tab 4 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab 6 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 8 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 7 \tab - \tab - \tab 1 \tab 6 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 2 \tab 1 \tab 25 \tab 1 \tab 9 \tab 8 \tab 31 \tab 25 \tab - \tab 2 \tab 9 \tab 3 \tab 4 \tab 1 \tab 17 \tab 1 \tab - \tab 1 \tab 35 \tab 3 \tab 4 \tab - \tab 6 \tab 38 \tab 6 \tab 8 \tab - \tab 1 \tab 1 \tab 40 \tab 27 \tab 1 \tab 4 \tab - \tab 2 \tab 2 \tab 29 \tab 5 \tab 14 \tab 8 \tab 2 \tab 43 \tab 16 \tab 11 \tab 5 \tab 16 \tab 8 \tab 12 \tab 5 \tab 6 \tab - \tab 15 \tab 6 \tab 5 \tab - \tab 1 \tab - \tab - \tab 4 \tab 16 \tab 8 \tab 25 \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab 6 \tab 3 \tab 7 \tab - \tab - \tab 7 \tab - \tab 39 \tab 13 \tab 1 \tab 4 \tab - \tab 3 \tab 9 \tab 3 \tab 5 \tab 2 \tab 6 \tab 18 \tab 9 \tab 1 \tab - \tab - \tab 11 \tab 2 \tab - \tab 32 \tab - \tab 8 \tab 1 \tab 24 \tab 63 \tab 2 \tab 1 \tab 5 \tab 2 \tab 44 \tab 13 \tab 2 \tab 11 \tab 12 \tab 2 \tab 3 \tab 8 \tab 32 \tab 2 \tab 1 \tab 3 \tab 15 \tab 8 \tab 9 \tab 12 \tab 12 \tab 3 \tab - \tab 3 \tab 3 \tab 5 \tab 2 \tab 2 \tab 1 \tab 9 \tab 28 \tab 2 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 5 \tab 10 \tab 1 \tab - \tab 3 \tab 14 \tab - \tab 5 \tab 12 \tab 8 \tab - \tab - \tab 6 \tab 26 \tab 7 \tab 5 \tab 1 \tab 5 \tab 5 \tab 12 \tab 17 \tab 1 \tab 8 \tab 2 \tab 8 \tab 15 \tab 2 \tab 1 \tab 2 \tab - \tab 17 \tab 2 \tab 32 \tab 5 \tab 3 \tab 25 \tab 16 \tab 3 \tab 4 \tab 1 \tab 2 \tab 9 \tab 16 \tab 3 \tab 3 \tab - \tab 15 \tab - \tab 9 \tab 14 \tab - \tab 9 \tab 13 \tab 3 \tab 47 \tab 3 \tab 5 \tab 1 \tab 1 \tab 16 \tab 16 \tab 2 \tab 30 \tab 24 \tab 1 \tab 4 \tab - \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab 10 \tab - \tab 3 \tab 2 \tab 117 \tab 1 \tab 18 \tab - \tab - \tab 2 \tab 2 \tab 5 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 5 \tab 2 \tab - \tab - \tab - \tab 6 \tab - \tab 2 \tab 4 \tab - \tab 7 \tab 3 \tab 3 \tab - \tab 9 \tab 8 \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab 3 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 6 \tab 4 \tab 10 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 5 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 4 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 4 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 3 \tab - \tab 1 \tab 31 \tab - \tab 6 \tab 3 \tab 40 \tab 17 \tab - \tab 2 \tab 6 \tab 5 \tab 6 \tab 4 \tab 24 \tab 3 \tab 1 \tab 3 \tab 25 \tab 3 \tab 2 \tab - \tab 2 \tab 42 \tab 7 \tab 4 \tab 1 \tab - \tab 2 \tab 34 \tab 28 \tab - \tab 3 \tab 2 \tab 5 \tab 1 \tab 24 \tab 8 \tab 12 \tab 4 \tab 1 \tab 43 \tab 20 \tab 9 \tab 5 \tab 26 \tab 14 \tab 20 \tab 1 \tab 3 \tab - \tab 26 \tab 6 \tab 11 \tab - \tab 1 \tab - \tab 3 \tab 4 \tab 4 \tab 9 \tab 18 \tab - \tab 4 \tab 1 \tab - \tab 2 \tab 4 \tab 2 \tab 8 \tab 1 \tab - \tab 7 \tab - \tab 39 \tab 6 \tab 1 \tab 7 \tab - \tab - \tab 9 \tab 2 \tab 1 \tab 3 \tab 1 \tab 15 \tab 9 \tab 4 \tab 1 \tab - \tab 8 \tab - \tab 1 \tab 40 \tab - \tab 8 \tab 1 \tab 20 \tab 58 \tab - \tab - \tab 4 \tab 2 \tab 35 \tab 6 \tab - \tab 9 \tab 13 \tab 5 \tab 2 \tab 5 \tab 32 \tab 1 \tab 1 \tab 1 \tab 10 \tab 7 \tab 5 \tab 10 \tab 6 \tab 1 \tab - \tab 2 \tab 2 \tab 5 \tab 1 \tab 5 \tab 3 \tab - \tab 23 \tab 3 \tab 1 \tab 5 \tab - \tab - \tab - \tab 3 \tab 5 \tab 11 \tab - \tab 1 \tab 6 \tab 10 \tab 1 \tab 2 \tab 7 \tab 10 \tab 2 \tab - \tab 4 \tab 14 \tab 9 \tab 2 \tab - \tab 6 \tab 3 \tab 7 \tab 11 \tab 1 \tab 6 \tab 1 \tab 11 \tab 11 \tab 3 \tab 2 \tab 5 \tab - \tab 16 \tab - \tab 39 \tab 3 \tab 4 \tab 23 \tab 19 \tab 1 \tab - \tab 2 \tab 2 \tab 9 \tab 12 \tab 2 \tab 5 \tab 2 \tab 8 \tab 1 \tab 15 \tab 25 \tab 1 \tab 2 \tab 9 \tab 9 \tab 41 \tab 4 \tab 5 \tab 1 \tab 1 \tab 29 \tab 11 \tab 3 \tab 46 \tab 24 \tab 3 \tab 7 \tab - \tab 16 \tab 1 \tab 2 \tab - \tab - \tab 8 \tab 1 \tab - \tab 7 \tab 103 \tab 3 \tab 17 \tab 1 \tab - \tab 3 \tab 5 \tab 3 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 5 \tab 2 \tab 5 \tab - \tab 1 \tab 1 \tab 5 \tab 3 \tab 1 \tab 2 \tab 1 \tab 6 \tab 2 \tab 20 \tab - \tab 9 \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 5 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab - \tab 3 \tab 2 \tab 7 \tab 3 \tab 10 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2 \tab - \tab 1 \tab 42 \tab - \tab 7 \tab 8 \tab 40 \tab 35 \tab 3 \tab 4 \tab 8 \tab 3 \tab 6 \tab - \tab 15 \tab 4 \tab - \tab 1 \tab 35 \tab 4 \tab 1 \tab - \tab 6 \tab 42 \tab 8 \tab 7 \tab 2 \tab - \tab 1 \tab 44 \tab 19 \tab 4 \tab 3 \tab 5 \tab 2 \tab - \tab 33 \tab 9 \tab 26 \tab 9 \tab 8 \tab 72 \tab 25 \tab 17 \tab 6 \tab 28 \tab 16 \tab 11 \tab 3 \tab 3 \tab - \tab 31 \tab 14 \tab 14 \tab 3 \tab 1 \tab - \tab 1 \tab 5 \tab 5 \tab 8 \tab 21 \tab 1 \tab 4 \tab 1 \tab - \tab 5 \tab 6 \tab 1 \tab 17 \tab 1 \tab - \tab 11 \tab 4 \tab 50 \tab 6 \tab - \tab 8 \tab - \tab 2 \tab 5 \tab 4 \tab 2 \tab 2 \tab 7 \tab 31 \tab 10 \tab 4 \tab - \tab - \tab 16 \tab - \tab - \tab 36 \tab 1 \tab 13 \tab 4 \tab 33 \tab 66 \tab 2 \tab 5 \tab 10 \tab 2 \tab 58 \tab 14 \tab - \tab 12 \tab 14 \tab - \tab 2 \tab 7 \tab 50 \tab 3 \tab 4 \tab - \tab 9 \tab 9 \tab 6 \tab 7 \tab 15 \tab 1 \tab 1 \tab 3 \tab 2 \tab 4 \tab 4 \tab 6 \tab 1 \tab 22 \tab 40 \tab 5 \tab 10 \tab 7 \tab 4 \tab 1 \tab - \tab 5 \tab 3 \tab 9 \tab - \tab 4 \tab 4 \tab 22 \tab - \tab 12 \tab 10 \tab 9 \tab 1 \tab - \tab 4 \tab 29 \tab 6 \tab 3 \tab 1 \tab 5 \tab 8 \tab 11 \tab 15 \tab 3 \tab 8 \tab 2 \tab 16 \tab 16 \tab 4 \tab 1 \tab 4 \tab - \tab 27 \tab - \tab 27 \tab 9 \tab 7 \tab 23 \tab 34 \tab 2 \tab 2 \tab 3 \tab 7 \tab 7 \tab 9 \tab 3 \tab 4 \tab 6 \tab 21 \tab 1 \tab 8 \tab 29 \tab 1 \tab 6 \tab 19 \tab 8 \tab 51 \tab 4 \tab 5 \tab 4 \tab 2 \tab 29 \tab 15 \tab 1 \tab 31 \tab 22 \tab 7 \tab 7 \tab - \tab 15 \tab 8 \tab 2 \tab 2 \tab 1 \tab 15 \tab 1 \tab 2 \tab 13 \tab 111 \tab 2 \tab 9 \tab 1 \tab - \tab 1 \tab 7 \tab 3 \tab 1 \tab 4 \tab 3 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 10 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 4 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 4 \tab 2 \tab 10 \tab 1 \tab 1 \tab 1 \tab 13 \tab 5 \tab - \tab - \tab - \tab 4 \tab - \tab 5 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 7 \tab 3 \tab 2 \tab - \tab 6 \tab 4 \tab 2 \tab 6 \tab 5 \tab 5 \tab - \tab 4 \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab - \tab - \tab 4 \tab 2 \tab - \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab 3 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 6 \tab 1 \tab 5 \tab 4 \tab 1 \tab - \tab - \tab 10 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 6 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab 2538 \cr
   2016 \tab 7 \tab - \tab 5 \tab 34 \tab - \tab 6 \tab 5 \tab 49 \tab 30 \tab - \tab 5 \tab 5 \tab 3 \tab 7 \tab 4 \tab 22 \tab 6 \tab - \tab 1 \tab 37 \tab 7 \tab 2 \tab - \tab 6 \tab 49 \tab 8 \tab 10 \tab 1 \tab - \tab 1 \tab 44 \tab 32 \tab 1 \tab 1 \tab 7 \tab 4 \tab - \tab 31 \tab 3 \tab 13 \tab 10 \tab 9 \tab 72 \tab 23 \tab 17 \tab 6 \tab 34 \tab 18 \tab 24 \tab 1 \tab 8 \tab - \tab 32 \tab 7 \tab 18 \tab 1 \tab 3 \tab - \tab 2 \tab 2 \tab 16 \tab 13 \tab 30 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 10 \tab 5 \tab 17 \tab 5 \tab 1 \tab 15 \tab 3 \tab 54 \tab 25 \tab - \tab 8 \tab - \tab 2 \tab 12 \tab 5 \tab 3 \tab 2 \tab 4 \tab 29 \tab 6 \tab 9 \tab 2 \tab - \tab 15 \tab - \tab - \tab 58 \tab - \tab 15 \tab 5 \tab 39 \tab 73 \tab - \tab 1 \tab 5 \tab 4 \tab 54 \tab 9 \tab - \tab 8 \tab 20 \tab 5 \tab 2 \tab 7 \tab 53 \tab 5 \tab - \tab 3 \tab 7 \tab 15 \tab 11 \tab 14 \tab 18 \tab 4 \tab - \tab 1 \tab 1 \tab 5 \tab 3 \tab 5 \tab 5 \tab 10 \tab 31 \tab 4 \tab 4 \tab 4 \tab 2 \tab 1 \tab - \tab 4 \tab 3 \tab 12 \tab - \tab 4 \tab 4 \tab 5 \tab 1 \tab 6 \tab 10 \tab 12 \tab 2 \tab 3 \tab 9 \tab 31 \tab 10 \tab 4 \tab 1 \tab 10 \tab 4 \tab 21 \tab 17 \tab 5 \tab 7 \tab 1 \tab 23 \tab 21 \tab 5 \tab 1 \tab 5 \tab - \tab 28 \tab 2 \tab 51 \tab 9 \tab 6 \tab 21 \tab 24 \tab 10 \tab 2 \tab 3 \tab 4 \tab 6 \tab 18 \tab 7 \tab 4 \tab 5 \tab 10 \tab 4 \tab 20 \tab 26 \tab 3 \tab 11 \tab 24 \tab 14 \tab 54 \tab 6 \tab 3 \tab 3 \tab 4 \tab 34 \tab 11 \tab 8 \tab 50 \tab 35 \tab 8 \tab 5 \tab 2 \tab 20 \tab 6 \tab 2 \tab - \tab 1 \tab 14 \tab 1 \tab 1 \tab 11 \tab 119 \tab 4 \tab 38 \tab 1 \tab - \tab 4 \tab 6 \tab 5 \tab - \tab 3 \tab 4 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 13 \tab 1 \tab 4 \tab 3 \tab - \tab 3 \tab 2 \tab 3 \tab 3 \tab 1 \tab 2 \tab 2 \tab 15 \tab 4 \tab 1 \tab 4 \tab 1 \tab 11 \tab 2 \tab 3 \tab 1 \tab 13 \tab 3 \tab - \tab - \tab 1 \tab 5 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab 3 \tab 3 \tab - \tab - \tab 4 \tab 2 \tab 4 \tab 14 \tab 5 \tab 4 \tab 1 \tab 7 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 6 \tab 1 \tab - \tab 4 \tab 1 \tab 7 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab 2 \tab 2 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 3 \tab 3 \tab 4 \tab 1 \tab 5 \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab 3 \tab 4 \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab - \tab - \tab 7 \tab 2 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 2867 \cr
   2018 \tab 3 \tab - \tab - \tab 33 \tab - \tab 8 \tab 1 \tab 44 \tab 31 \tab 3 \tab 5 \tab 12 \tab 2 \tab 14 \tab 7 \tab 20 \tab 2 \tab - \tab 3 \tab 28 \tab 6 \tab 2 \tab - \tab 1 \tab 37 \tab 13 \tab 8 \tab 1 \tab 1 \tab - \tab 27 \tab 21 \tab 1 \tab 3 \tab 5 \tab 4 \tab 2 \tab 38 \tab 12 \tab 25 \tab 2 \tab 6 \tab 60 \tab 13 \tab 12 \tab 5 \tab 23 \tab 20 \tab 13 \tab 1 \tab 2 \tab 1 \tab 16 \tab 2 \tab 14 \tab - \tab 4 \tab - \tab 3 \tab 5 \tab 9 \tab 8 \tab 23 \tab - \tab 4 \tab - \tab 2 \tab 3 \tab 3 \tab 1 \tab 8 \tab 3 \tab - \tab 14 \tab 2 \tab 39 \tab 17 \tab 2 \tab 7 \tab 1 \tab 5 \tab 9 \tab 1 \tab 1 \tab 2 \tab 6 \tab 18 \tab 7 \tab 9 \tab 2 \tab - \tab 18 \tab - \tab - \tab 18 \tab - \tab 16 \tab 6 \tab 14 \tab 61 \tab 2 \tab 2 \tab 15 \tab 3 \tab 44 \tab 12 \tab 1 \tab 9 \tab 9 \tab 5 \tab 2 \tab 5 \tab 29 \tab 1 \tab 1 \tab 1 \tab 6 \tab 12 \tab 18 \tab 5 \tab 11 \tab 1 \tab - \tab 1 \tab - \tab 4 \tab - \tab 11 \tab 1 \tab 9 \tab 26 \tab 3 \tab 6 \tab 5 \tab 1 \tab - \tab 1 \tab 7 \tab 2 \tab 7 \tab - \tab 3 \tab 6 \tab 17 \tab 1 \tab 12 \tab 10 \tab 14 \tab 3 \tab 5 \tab 6 \tab 32 \tab 13 \tab 3 \tab 1 \tab 9 \tab 10 \tab 16 \tab 14 \tab - \tab 4 \tab 1 \tab 15 \tab 20 \tab 7 \tab 3 \tab 10 \tab 2 \tab 24 \tab 3 \tab 27 \tab 3 \tab 4 \tab 24 \tab 27 \tab 6 \tab 1 \tab 4 \tab - \tab 8 \tab 9 \tab 3 \tab 1 \tab 6 \tab 25 \tab 1 \tab 9 \tab 23 \tab 3 \tab 4 \tab 22 \tab 9 \tab 40 \tab 4 \tab 2 \tab 2 \tab 1 \tab 26 \tab 9 \tab 8 \tab 53 \tab 30 \tab 5 \tab 4 \tab 1 \tab 14 \tab 3 \tab 4 \tab 1 \tab - \tab 19 \tab - \tab 4 \tab 5 \tab 100 \tab 3 \tab 14 \tab 2 \tab - \tab 2 \tab 12 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 6 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 3 \tab 6 \tab 2 \tab 5 \tab 2 \tab 4 \tab 1 \tab 2 \tab - \tab 8 \tab 9 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab 4 \tab - \tab - \tab 2 \tab 1 \tab 7 \tab 1 \tab 1 \tab - \tab 5 \tab 2 \tab 3 \tab 10 \tab 13 \tab - \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab 4 \tab 4 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab 2 \tab 9 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 6 \tab - \tab 1 \tab 35 \tab 1 \tab 3 \tab 3 \tab 55 \tab 21 \tab 1 \tab 3 \tab 6 \tab 2 \tab 4 \tab 6 \tab 16 \tab 6 \tab - \tab 2 \tab 32 \tab 5 \tab 3 \tab - \tab 9 \tab 65 \tab 11 \tab 6 \tab 3 \tab - \tab 2 \tab 55 \tab 23 \tab - \tab 3 \tab 6 \tab 1 \tab - \tab 41 \tab 5 \tab 19 \tab 8 \tab 5 \tab 40 \tab 19 \tab 15 \tab 3 \tab 30 \tab 35 \tab 19 \tab 2 \tab 8 \tab - \tab 44 \tab 22 \tab 14 \tab 2 \tab 2 \tab - \tab 3 \tab 1 \tab 13 \tab 15 \tab 24 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 6 \tab 3 \tab 15 \tab 1 \tab - \tab 12 \tab 5 \tab 64 \tab 18 \tab 3 \tab 7 \tab - \tab 1 \tab 19 \tab 4 \tab 4 \tab 1 \tab 10 \tab 55 \tab 8 \tab 12 \tab 2 \tab 2 \tab 16 \tab 1 \tab 2 \tab 73 \tab - \tab 18 \tab 11 \tab 51 \tab 98 \tab 2 \tab - \tab 16 \tab 2 \tab 73 \tab 19 \tab 1 \tab 20 \tab 25 \tab 4 \tab 12 \tab 10 \tab 61 \tab 5 \tab 4 \tab 3 \tab 21 \tab 16 \tab 10 \tab 6 \tab 22 \tab 7 \tab 1 \tab 3 \tab 2 \tab 6 \tab 3 \tab 8 \tab 1 \tab 42 \tab 39 \tab 9 \tab 4 \tab 4 \tab 4 \tab 3 \tab - \tab 6 \tab 4 \tab 21 \tab 1 \tab 1 \tab 7 \tab 33 \tab - \tab 16 \tab 30 \tab 21 \tab - \tab 2 \tab 12 \tab 34 \tab 10 \tab 9 \tab 1 \tab 8 \tab 11 \tab 26 \tab 36 \tab 5 \tab 24 \tab 3 \tab 35 \tab 45 \tab 6 \tab 4 \tab 7 \tab - \tab 42 \tab 1 \tab 108 \tab 7 \tab 8 \tab 31 \tab 52 \tab 7 \tab 4 \tab 10 \tab 10 \tab 17 \tab 23 \tab 5 \tab 7 \tab 6 \tab 30 \tab 4 \tab 49 \tab 58 \tab 4 \tab 15 \tab 22 \tab 18 \tab 132 \tab 16 \tab 12 \tab 4 \tab 4 \tab 50 \tab 37 \tab 8 \tab 95 \tab 70 \tab 16 \tab 13 \tab 1 \tab 47 \tab 12 \tab 8 \tab 3 \tab - \tab 30 \tab - \tab 3 \tab 24 \tab 159 \tab 4 \tab 54 \tab 1 \tab 1 \tab 1 \tab 8 \tab 2 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 7 \tab 6 \tab 7 \tab 3 \tab 1 \tab 3 \tab 7 \tab 2 \tab 4 \tab 1 \tab 6 \tab 1 \tab 13 \tab 3 \tab 2 \tab 14 \tab - \tab 17 \tab 4 \tab - \tab 2 \tab 28 \tab 7 \tab - \tab 2 \tab - \tab 17 \tab 2 \tab 4 \tab 8 \tab 4 \tab 4 \tab - \tab 4 \tab - \tab 4 \tab 2 \tab 3 \tab 2 \tab 3 \tab 4 \tab 4 \tab 12 \tab 12 \tab - \tab 2 \tab 21 \tab 6 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab 3 \tab 4 \tab - \tab - \tab 3 \tab 3 \tab 3 \tab - \tab 2 \tab 1 \tab 4 \tab 7 \tab 8 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 3 \tab 9 \tab 3 \tab 4 \tab 1 \tab 3 \tab 5 \tab - \tab 1 \tab 1 \tab 5 \tab 1 \tab 7 \tab 7 \tab - \tab 1 \tab 1 \tab 4 \tab 4 \tab - \tab 14 \tab 3 \tab - \tab 3 \tab 1 \tab 7 \tab 3 \tab 1 \tab 4 \tab 2 \tab 6 \tab 3 \tab 3 \tab 2 \tab 1 \tab - \tab 7 \tab 4 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 7 \tab - \tab - \tab 3 \tab 4032 \cr
   2022 \tab 5 \tab - \tab - \tab 35 \tab - \tab 9 \tab 6 \tab 62 \tab 24 \tab - \tab 2 \tab 8 \tab 4 \tab 9 \tab 12 \tab 24 \tab 6 \tab - \tab 2 \tab 34 \tab 3 \tab 1 \tab 3 \tab 6 \tab 43 \tab 5 \tab 14 \tab 3 \tab 1 \tab - \tab 57 \tab 26 \tab - \tab 2 \tab 3 \tab 4 \tab 1 \tab 45 \tab 9 \tab 24 \tab 10 \tab 4 \tab 61 \tab 31 \tab 27 \tab 7 \tab 53 \tab 26 \tab 9 \tab 5 \tab 7 \tab 1 \tab 37 \tab 10 \tab 9 \tab - \tab 4 \tab - \tab 2 \tab 4 \tab 9 \tab 9 \tab 47 \tab - \tab - \tab - \tab 1 \tab 1 \tab 9 \tab 3 \tab 24 \tab - \tab 1 \tab 16 \tab 5 \tab 71 \tab 17 \tab 2 \tab 6 \tab - \tab 3 \tab 19 \tab 4 \tab 1 \tab 1 \tab 10 \tab 38 \tab 14 \tab 11 \tab 2 \tab - \tab 12 \tab - \tab 1 \tab 71 \tab - \tab 15 \tab 6 \tab 38 \tab 78 \tab - \tab - \tab 11 \tab 1 \tab 64 \tab 29 \tab 3 \tab 18 \tab 19 \tab 4 \tab 2 \tab 14 \tab 65 \tab 10 \tab 3 \tab 3 \tab 9 \tab 19 \tab 19 \tab 14 \tab 20 \tab - \tab - \tab - \tab 2 \tab 4 \tab 3 \tab 11 \tab 8 \tab 21 \tab 19 \tab 2 \tab 7 \tab 8 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 16 \tab 4 \tab 5 \tab 5 \tab 18 \tab 3 \tab 13 \tab 23 \tab 18 \tab 1 \tab 2 \tab 12 \tab 41 \tab 9 \tab 8 \tab - \tab 10 \tab 16 \tab 26 \tab 24 \tab 5 \tab 10 \tab 1 \tab 24 \tab 28 \tab 3 \tab 2 \tab 7 \tab - \tab 42 \tab - \tab 72 \tab 11 \tab 7 \tab 34 \tab 45 \tab 7 \tab 4 \tab 5 \tab 3 \tab 10 \tab 12 \tab 3 \tab 2 \tab 8 \tab 20 \tab 2 \tab 24 \tab 32 \tab 6 \tab 4 \tab 22 \tab 11 \tab 104 \tab 10 \tab 8 \tab 2 \tab 1 \tab 28 \tab 22 \tab 5 \tab 86 \tab 28 \tab 14 \tab 6 \tab 3 \tab 38 \tab 4 \tab 3 \tab 2 \tab - \tab 23 \tab - \tab 2 \tab 21 \tab 218 \tab 7 \tab 48 \tab - \tab - \tab 5 \tab 8 \tab 3 \tab - \tab 1 \tab 3 \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 4 \tab 1 \tab 7 \tab 1 \tab 5 \tab 1 \tab 3 \tab 3 \tab 8 \tab 2 \tab 2 \tab - \tab 4 \tab 3 \tab 5 \tab 5 \tab 1 \tab 6 \tab 3 \tab - \tab 1 \tab 17 \tab 7 \tab - \tab 1 \tab - \tab 13 \tab 3 \tab 3 \tab 5 \tab 4 \tab - \tab - \tab 6 \tab 3 \tab 5 \tab 4 \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 10 \tab 7 \tab - \tab - \tab 13 \tab 3 \tab 3 \tab - \tab 8 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 8 \tab 4 \tab 3 \tab 1 \tab - \tab 3 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 4 \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 8 \tab 3 \tab 8 \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 7 \tab - \tab - \tab 4 \tab - \tab 7 \tab 1 \tab 9 \tab 2 \tab - \tab - \tab 2 \tab 4 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 5 \tab 3 \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 6 \tab 2 \tab 1 \tab - \tab 3544 \cr
   Total \tab 107 \tab 80 \tab 92 \tab 1209 \tab 53 \tab 573 \tab 132 \tab 1068 \tab 691 \tab 53 \tab 86 \tab 265 \tab 143 \tab 191 \tab 100 \tab 403 \tab 169 \tab 8 \tab 56 \tab 1077 \tab 143 \tab 77 \tab 42 \tab 121 \tab 1070 \tab 261 \tab 215 \tab 70 \tab 50 \tab 42 \tab 1110 \tab 634 \tab 46 \tab 49 \tab 146 \tab 73 \tab 55 \tab 941 \tab 253 \tab 560 \tab 205 \tab 97 \tab 1385 \tab 629 \tab 482 \tab 190 \tab 580 \tab 501 \tab 427 \tab 101 \tab 186 \tab 17 \tab 817 \tab 220 \tab 341 \tab 44 \tab 48 \tab 41 \tab 80 \tab 102 \tab 364 \tab 232 \tab 809 \tab 34 \tab 51 \tab 8 \tab 63 \tab 78 \tab 208 \tab 173 \tab 400 \tab 61 \tab 12 \tab 307 \tab 178 \tab 1379 \tab 464 \tab 42 \tab 200 \tab 28 \tab 87 \tab 299 \tab 72 \tab 279 \tab 52 \tab 234 \tab 725 \tab 289 \tab 134 \tab 64 \tab 21 \tab 251 \tab 27 \tab 14 \tab 679 \tab 18 \tab 552 \tab 120 \tab 1297 \tab 2924 \tab 39 \tab 67 \tab 209 \tab 87 \tab 1516 \tab 720 \tab 30 \tab 352 \tab 302 \tab 139 \tab 191 \tab 134 \tab 1276 \tab 72 \tab 67 \tab 56 \tab 267 \tab 339 \tab 368 \tab 569 \tab 485 \tab 119 \tab 9 \tab 83 \tab 68 \tab 147 \tab 56 \tab 202 \tab 66 \tab 264 \tab 811 \tab 126 \tab 214 \tab 109 \tab 34 \tab 25 \tab 15 \tab 93 \tab 91 \tab 251 \tab 20 \tab 100 \tab 92 \tab 397 \tab 21 \tab 275 \tab 365 \tab 305 \tab 36 \tab 69 \tab 172 \tab 595 \tab 246 \tab 151 \tab 37 \tab 174 \tab 129 \tab 371 \tab 588 \tab 61 \tab 223 \tab 99 \tab 391 \tab 397 \tab 86 \tab 41 \tab 126 \tab 14 \tab 546 \tab 26 \tab 1007 \tab 115 \tab 148 \tab 437 \tab 560 \tab 65 \tab 37 \tab 111 \tab 68 \tab 174 \tab 237 \tab 78 \tab 81 \tab 85 \tab 403 \tab 75 \tab 356 \tab 627 \tab 50 \tab 139 \tab 388 \tab 185 \tab 1475 \tab 213 \tab 136 \tab 34 \tab 36 \tab 870 \tab 293 \tab 109 \tab 1114 \tab 663 \tab 166 \tab 191 \tab 42 \tab 302 \tab 97 \tab 62 \tab 26 \tab 47 \tab 364 \tab 14 \tab 58 \tab 232 \tab 5142 \tab 69 \tab 342 \tab 35 \tab 9 \tab 40 \tab 123 \tab 68 \tab 31 \tab 69 \tab 62 \tab 23 \tab 44 \tab 106 \tab 44 \tab 21 \tab 39 \tab 12 \tab 19 \tab 19 \tab 82 \tab 17 \tab 27 \tab 122 \tab 72 \tab 50 \tab 63 \tab 27 \tab 29 \tab 93 \tab 118 \tab 65 \tab 26 \tab 64 \tab 27 \tab 139 \tab 43 \tab 31 \tab 106 \tab 28 \tab 142 \tab 44 \tab 70 \tab 35 \tab 235 \tab 100 \tab 2 \tab 27 \tab 21 \tab 93 \tab 61 \tab 44 \tab 57 \tab 59 \tab 16 \tab 13 \tab 43 \tab 23 \tab 125 \tab 71 \tab 30 \tab 6 \tab 56 \tab 72 \tab 53 \tab 161 \tab 144 \tab 122 \tab 17 \tab 68 \tab 53 \tab 27 \tab 18 \tab 54 \tab 8 \tab 46 \tab 28 \tab 5 \tab 40 \tab 24 \tab 8 \tab 21 \tab 24 \tab 73 \tab 81 \tab 15 \tab 28 \tab 21 \tab 28 \tab 26 \tab 70 \tab 12 \tab 24 \tab 22 \tab 69 \tab 26 \tab 8 \tab 23 \tab 41 \tab 15 \tab 24 \tab 14 \tab 31 \tab 32 \tab 51 \tab 6 \tab 8 \tab 14 \tab 40 \tab 32 \tab 47 \tab 73 \tab 23 \tab 64 \tab 12 \tab 47 \tab 60 \tab 12 \tab 13 \tab 14 \tab 56 \tab 18 \tab 23 \tab 33 \tab 11 \tab 14 \tab 35 \tab 33 \tab 57 \tab 18 \tab 42 \tab 24 \tab 10 \tab 9 \tab 16 \tab 61 \tab 37 \tab 46 \tab 16 \tab 8 \tab 22 \tab 67 \tab 31 \tab 16 \tab 9 \tab 10 \tab 40 \tab 9 \tab 3 \tab 2 \tab 10 \tab 8 \tab 8 \tab 6 \tab 30 \tab 7 \tab 2 \tab 5 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} i don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
