% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_83.R
\name{sibdeath}
\alias{sibdeath}
\title{Death of sibling}
\source{
General Social Survey https://gss.norc.org
}
\description{
sibdeath
}
\details{
Question 275d. Death of sibling or sibling-in-law.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab died 1-5 yrs ago \tab died 6+ yrs ago \tab died last year \tab no answer \tab no death reported \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 98 \tab 166 \tab 201 \tab 75 \tab 20 \tab 972 \tab 1532 \cr
   1980 \tab 73 \tab 151 \tab 193 \tab 78 \tab 11 \tab 962 \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 98 \tab 181 \tab 209 \tab 69 \tab 7 \tab 1035 \tab 1599 \cr
   1984 \tab 64 \tab 130 \tab 227 \tab 69 \tab 21 \tab 962 \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 60 \tab 156 \tab 221 \tab 84 \tab 6 \tab 943 \tab 1470 \cr
   1987 \tab 89 \tab 177 \tab 250 \tab 105 \tab 20 \tab 1178 \tab 1819 \cr
   1988 \tab 528 \tab 114 \tab 133 \tab 69 \tab 6 \tab 631 \tab 1481 \cr
   1989 \tab 549 \tab 73 \tab 115 \tab 36 \tab 9 \tab 755 \tab 1537 \cr
   1990 \tab 516 \tab 56 \tab 97 \tab 26 \tab 7 \tab 670 \tab 1372 \cr
   1991 \tab 550 \tab 59 \tab 129 \tab 32 \tab 15 \tab 732 \tab 1517 \cr
   1993 \tab 616 \tab 72 \tab 104 \tab 30 \tab 12 \tab 772 \tab 1606 \cr
   1994 \tab 2510 \tab 35 \tab 53 \tab 11 \tab 3 \tab 380 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 58275 \tab 1370 \tab 1932 \tab 684 \tab 137 \tab 9992 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} no death reported
\item \code{1} died 6+ yrs ago
\item \code{2} died 1-5 yrs ago
\item \code{3} died last year
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
