% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_09.R
\name{sibs}
\alias{sibs}
\title{Number of brothers and sisters}
\source{
General Social Survey https://gss.norc.org
}
\description{
sibs
}
\details{
Question 11. How many brothers and sisters did you have? Please count those born alive, but no longer living, as well as those alive now. Also include stepbrothers and stepsisters, and children adopted by your parents.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab don't know \tab no answer \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 18 \tab 20 \tab 27 \tab 8 \tab 9 \tab 17 \tab 19 \tab 21 \tab 31 \tab 22 \tab 26 \tab 23 \tab 24 \tab 68 \tab 25 \tab 35 \tab 30 \tab 32 \tab 33 \tab 29 \tab 34 \tab iap \tab 37 \tab 55 \tab 28 \tab 43 \tab skipped on web \tab 36 \tab Total \cr
   1972 \tab 103 \tab 249 \tab 249 \tab 217 \tab 169 \tab 130 \tab 107 \tab 382 \tab 1 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 86 \tab 221 \tab 231 \tab 197 \tab 159 \tab 129 \tab 112 \tab 81 \tab 3 \tab - \tab 36 \tab 44 \tab 25 \tab 14 \tab 10 \tab 7 \tab 2 \tab 2 \tab 1 \tab 1 \tab 65 \tab 78 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 76 \tab 218 \tab 214 \tab 224 \tab 181 \tab 131 \tab 94 \tab 92 \tab 1 \tab - \tab 39 \tab 21 \tab 27 \tab 14 \tab 5 \tab 6 \tab 4 \tab 1 \tab 1 \tab 1 \tab 76 \tab 55 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 95 \tab 206 \tab 243 \tab 212 \tab 145 \tab 125 \tab 110 \tab 93 \tab 2 \tab 3 \tab 40 \tab 30 \tab 18 \tab 15 \tab 7 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab 75 \tab 60 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 89 \tab 194 \tab 222 \tab 222 \tab 175 \tab 153 \tab 98 \tab 98 \tab - \tab 1 \tab 47 \tab 30 \tab 22 \tab 10 \tab 5 \tab 8 \tab 3 \tab - \tab - \tab - \tab 64 \tab 54 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 80 \tab 194 \tab 249 \tab 188 \tab 183 \tab 138 \tab 124 \tab 105 \tab 1 \tab 2 \tab 40 \tab 29 \tab 21 \tab 15 \tab 9 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 75 \tab 67 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 98 \tab 216 \tab 268 \tab 210 \tab 168 \tab 133 \tab 131 \tab 91 \tab - \tab - \tab 37 \tab 22 \tab 23 \tab 14 \tab 8 \tab 4 \tab 6 \tab - \tab 1 \tab - \tab 58 \tab 40 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 73 \tab 212 \tab 230 \tab 195 \tab 189 \tab 147 \tab 107 \tab 84 \tab 1 \tab - \tab 32 \tab 27 \tab 18 \tab 16 \tab 7 \tab 5 \tab 3 \tab 1 \tab 3 \tab 1 \tab 70 \tab 45 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 103 \tab 240 \tab 269 \tab 273 \tab 211 \tab 179 \tab 140 \tab 104 \tab 2 \tab 3 \tab 48 \tab 47 \tab 29 \tab 18 \tab 13 \tab 7 \tab - \tab 6 \tab 2 \tab - \tab 92 \tab 69 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 98 \tab 237 \tab 274 \tab 253 \tab 185 \tab 154 \tab 112 \tab 80 \tab 1 \tab - \tab 22 \tab 27 \tab 18 \tab 9 \tab 4 \tab 3 \tab 4 \tab - \tab 3 \tab - \tab 68 \tab 43 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 64 \tab 222 \tab 266 \tab 234 \tab 158 \tab 122 \tab 90 \tab 92 \tab - \tab 8 \tab 36 \tab 28 \tab 14 \tab 17 \tab 5 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 63 \tab 46 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 84 \tab 233 \tab 240 \tab 215 \tab 177 \tab 149 \tab 120 \tab 92 \tab 4 \tab 4 \tab 21 \tab 20 \tab 13 \tab 7 \tab 6 \tab 4 \tab 9 \tab 1 \tab 1 \tab - \tab 73 \tab 58 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 60 \tab 236 \tab 268 \tab 228 \tab 166 \tab 146 \tab 98 \tab 64 \tab - \tab 3 \tab 29 \tab 23 \tab 15 \tab 15 \tab 3 \tab 3 \tab 3 \tab - \tab - \tab - \tab 70 \tab 38 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 89 \tab 261 \tab 273 \tab 292 \tab 217 \tab 157 \tab 130 \tab 80 \tab 3 \tab 6 \tab 65 \tab 32 \tab 29 \tab 17 \tab 9 \tab 2 \tab 4 \tab - \tab 2 \tab 1 \tab 76 \tab 69 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 63 \tab 226 \tab 292 \tab 216 \tab 170 \tab 135 \tab 85 \tab 86 \tab 5 \tab 4 \tab 34 \tab 23 \tab 10 \tab 9 \tab 6 \tab 4 \tab 4 \tab 2 \tab - \tab - \tab 61 \tab 43 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 77 \tab 237 \tab 311 \tab 264 \tab 167 \tab 130 \tab 102 \tab 67 \tab - \tab 4 \tab 30 \tab 16 \tab 16 \tab 12 \tab 4 \tab 3 \tab 2 \tab - \tab - \tab - \tab 53 \tab 40 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 64 \tab 224 \tab 261 \tab 243 \tab 164 \tab 119 \tab 88 \tab 60 \tab - \tab 1 \tab 23 \tab 22 \tab 10 \tab 4 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab 49 \tab 31 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 74 \tab 236 \tab 276 \tab 236 \tab 209 \tab 118 \tab 80 \tab 81 \tab 4 \tab 8 \tab 34 \tab 22 \tab 11 \tab 9 \tab 5 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 58 \tab 47 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 89 \tab 260 \tab 341 \tab 282 \tab 170 \tab 117 \tab 93 \tab 67 \tab 3 \tab 3 \tab 23 \tab 15 \tab 11 \tab 11 \tab 3 \tab 4 \tab 2 \tab - \tab - \tab - \tab 60 \tab 47 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 162 \tab 537 \tab 581 \tab 483 \tab 354 \tab 232 \tab 185 \tab 131 \tab 5 \tab 6 \tab 30 \tab 33 \tab 28 \tab 20 \tab 16 \tab 8 \tab 5 \tab 1 \tab 2 \tab 2 \tab 89 \tab 76 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 140 \tab 501 \tab 570 \tab 497 \tab 350 \tab 228 \tab 154 \tab 136 \tab 2 \tab 5 \tab 47 \tab 50 \tab 20 \tab 17 \tab 5 \tab 8 \tab 7 \tab 1 \tab - \tab - \tab 84 \tab 53 \tab 1 \tab 2 \tab 13 \tab 5 \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 105 \tab 493 \tab 558 \tab 483 \tab 350 \tab 207 \tab 157 \tab 118 \tab 5 \tab 12 \tab 43 \tab 47 \tab 26 \tab 10 \tab 9 \tab 5 \tab 6 \tab 4 \tab 2 \tab - \tab 84 \tab 68 \tab 2 \tab 1 \tab 25 \tab - \tab 4 \tab - \tab 5 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 142 \tab 507 \tab 569 \tab 475 \tab 301 \tab 237 \tab 162 \tab 123 \tab 5 \tab 6 \tab 40 \tab 40 \tab 19 \tab 20 \tab 10 \tab 7 \tab 3 \tab - \tab 2 \tab - \tab 90 \tab 51 \tab 3 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 140 \tab 505 \tab 583 \tab 457 \tab 314 \tab 224 \tab 157 \tab 115 \tab 2 \tab 7 \tab 40 \tab 28 \tab 22 \tab 7 \tab 8 \tab 11 \tab 6 \tab 2 \tab 2 \tab - \tab 77 \tab 49 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 144 \tab 553 \tab 556 \tab 472 \tab 344 \tab 198 \tab 157 \tab 108 \tab 2 \tab 2 \tab 47 \tab 22 \tab 18 \tab 15 \tab 11 \tab 7 \tab 6 \tab 3 \tab 1 \tab - \tab 80 \tab 63 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 145 \tab 560 \tab 605 \tab 439 \tab 345 \tab 234 \tab 189 \tab 141 \tab 1 \tab 3 \tab 47 \tab 39 \tab 30 \tab 19 \tab 10 \tab 4 \tab 5 \tab 4 \tab 1 \tab - \tab 81 \tab 81 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1518 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 89 \tab 394 \tab 400 \tab 315 \tab 250 \tab 160 \tab 117 \tab 100 \tab - \tab 2 \tab 30 \tab 19 \tab 14 \tab 8 \tab 6 \tab 3 \tab 3 \tab 4 \tab 1 \tab - \tab 75 \tab 27 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 94 \tab 401 \tab 412 \tab 320 \tab 239 \tab 167 \tab 122 \tab 92 \tab 1 \tab 5 \tab 28 \tab 26 \tab 17 \tab 7 \tab 8 \tab 5 \tab 4 \tab - \tab - \tab - \tab 62 \tab 32 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 96 \tab 368 \tab 410 \tab 331 \tab 208 \tab 147 \tab 113 \tab 91 \tab 2 \tab 1 \tab 35 \tab 18 \tab 12 \tab 7 \tab 5 \tab 10 \tab 3 \tab 1 \tab 1 \tab - \tab 64 \tab 45 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 95 \tab 461 \tab 496 \tab 436 \tab 287 \tab 214 \tab 154 \tab 114 \tab 1 \tab 6 \tab 43 \tab 26 \tab 14 \tab 14 \tab 11 \tab 4 \tab 6 \tab 1 \tab 1 \tab - \tab 83 \tab 61 \tab 6 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 130 \tab 550 \tab 596 \tab 452 \tab 319 \tab 206 \tab 161 \tab 128 \tab 2 \tab 3 \tab 39 \tab 39 \tab 27 \tab 23 \tab 7 \tab 2 \tab 3 \tab 3 \tab 2 \tab 1 \tab 90 \tab 72 \tab 4 \tab - \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2867 \cr
   2018 \tab 95 \tab 472 \tab 483 \tab 384 \tab 244 \tab 176 \tab 147 \tab 125 \tab 4 \tab 1 \tab 34 \tab 19 \tab 13 \tab 14 \tab 5 \tab 2 \tab 4 \tab - \tab - \tab - \tab 76 \tab 44 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 216 \tab 886 \tab 935 \tab 679 \tab 406 \tab 287 \tab 186 \tab 116 \tab 49 \tab - \tab 29 \tab 33 \tab 25 \tab 12 \tab 9 \tab 6 \tab 1 \tab 1 \tab - \tab - \tab 81 \tab 55 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab 11 \tab - \tab 4032 \cr
   2022 \tab 174 \tab 743 \tab 790 \tab 565 \tab 400 \tab 243 \tab 175 \tab 110 \tab - \tab 12 \tab 49 \tab 33 \tab 27 \tab 15 \tab 8 \tab 7 \tab 6 \tab 2 \tab 1 \tab - \tab 90 \tab 65 \tab 5 \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 13 \tab 1 \tab 3544 \cr
   Total \tab 3532 \tab 12253 \tab 13521 \tab 11189 \tab 8074 \tab 5772 \tab 4357 \tab 3647 \tab 112 \tab 127 \tab 1217 \tab 950 \tab 642 \tab 434 \tab 242 \tab 161 \tab 119 \tab 47 \tab 32 \tab 7 \tab 2412 \tab 1772 \tab 58 \tab 28 \tab 53 \tab 6 \tab 19 \tab 7 \tab 16 \tab 6 \tab 1 \tab 9 \tab 2 \tab 4 \tab 4 \tab 3 \tab 2 \tab 1 \tab 1522 \tab 1 \tab 1 \tab 2 \tab 1 \tab 24 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
