% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_58.R
\name{socbar}
\alias{socbar}
\title{socbar}
\source{
General Social Survey https://gss.norc.org
}
\description{
Spend evening at bar
}
\details{
Question 173. Would you use this card and tell me which answer comes closest to how often you do the following things: d. Go to a bar or tavern
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab about once a month \tab about once a year \tab almost daily \tab don't know \tab never \tab no answer \tab once or twice a week \tab several times a month \tab several times a year \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 131 \tab 112 \tab 33 \tab 18 \tab 777 \tab 4 \tab 139 \tab 92 \tab 178 \tab - \tab 1484 \cr
   1975 \tab - \tab 136 \tab 140 \tab 22 \tab 10 \tab 832 \tab 4 \tab 114 \tab 98 \tab 134 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 152 \tab 144 \tab 39 \tab 1 \tab 763 \tab 4 \tab 139 \tab 124 \tab 164 \tab - \tab 1530 \cr
   1978 \tab - \tab 112 \tab 153 \tab 29 \tab 1 \tab 775 \tab 3 \tab 133 \tab 140 \tab 186 \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 162 \tab 193 \tab 30 \tab 9 \tab 935 \tab 7 \tab 178 \tab 122 \tab 224 \tab - \tab 1860 \cr
   1983 \tab - \tab 130 \tab 161 \tab 26 \tab 2 \tab 800 \tab 4 \tab 155 \tab 139 \tab 182 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 169 \tab 164 \tab 28 \tab 1 \tab 747 \tab 4 \tab 119 \tab 127 \tab 175 \tab - \tab 1534 \cr
   1986 \tab - \tab 142 \tab 145 \tab 16 \tab 2 \tab 782 \tab 5 \tab 118 \tab 88 \tab 172 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 97 \tab 123 \tab 10 \tab 1 \tab 501 \tab 4 \tab 82 \tab 65 \tab 105 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 85 \tab 109 \tab 10 \tab - \tab 559 \tab 2 \tab 82 \tab 56 \tab 103 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 94 \tab 106 \tab 6 \tab 1 \tab 479 \tab 3 \tab 74 \tab 59 \tab 106 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 109 \tab 119 \tab 18 \tab 2 \tab 539 \tab 4 \tab 70 \tab 62 \tab 101 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 110 \tab 120 \tab 7 \tab 1 \tab 556 \tab 3 \tab 85 \tab 78 \tab 120 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 180 \tab 243 \tab 20 \tab 5 \tab 971 \tab 10 \tab 142 \tab 145 \tab 261 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 198 \tab 250 \tab 25 \tab 8 \tab 948 \tab 1 \tab 144 \tab 136 \tab 250 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 199 \tab 239 \tab 26 \tab 3 \tab 916 \tab 7 \tab 125 \tab 124 \tab 232 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 180 \tab 233 \tab 20 \tab 11 \tab 922 \tab 6 \tab 142 \tab 140 \tab 223 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 114 \tab 119 \tab 11 \tab 2 \tab 401 \tab 1 \tab 78 \tab 74 \tab 108 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 91 \tab 142 \tab 6 \tab 1 \tab 432 \tab - \tab 65 \tab 61 \tab 108 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 200 \tab 259 \tab 14 \tab 1 \tab 1027 \tab 2 \tab 113 \tab 122 \tab 254 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 151 \tab 158 \tab 12 \tab 1 \tab 649 \tab - \tab 95 \tab 86 \tab 177 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 159 \tab 172 \tab 14 \tab 3 \tab 668 \tab 1 \tab 86 \tab 135 \tab 192 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 145 \tab 167 \tab 11 \tab 1 \tab 649 \tab 1 \tab 75 \tab 107 \tab 146 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 209 \tab 233 \tab 11 \tab - \tab 757 \tab - \tab 94 \tab 128 \tab 243 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 221 \tab 271 \tab 16 \tab 1 \tab 852 \tab 1 \tab 118 \tab 154 \tab 254 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 195 \tab 216 \tab 10 \tab 1 \tab 689 \tab 1 \tab 93 \tab 131 \tab 223 \tab - \tab 2348 \cr
   2021 \tab 1320 \tab 271 \tab 461 \tab 5 \tab - \tab 1171 \tab 1 \tab 91 \tab 190 \tab 513 \tab 9 \tab 4032 \cr
   2022 \tab 1168 \tab 272 \tab 361 \tab 11 \tab 3 \tab 1014 \tab 1 \tab 105 \tab 199 \tab 406 \tab 4 \tab 3544 \cr
   Total \tab 29103 \tab 4414 \tab 5313 \tab 486 \tab 90 \tab 21111 \tab 84 \tab 3054 \tab 3182 \tab 5540 \tab 13 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} almost daily
\item \code{2} once or twice a week
\item \code{3} several times a month
\item \code{4} about once a month
\item \code{5} several times a year
\item \code{6} about once a year
\item \code{7} never
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
