% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_58.R
\name{socfrend}
\alias{socfrend}
\title{socfrend}
\source{
General Social Survey https://gss.norc.org
}
\description{
Spend evening with friends
}
\details{
Question 173. Would you use this card and tell me which answer comes closest to how often you do the following things: c. Spend a social evening with friends who live outside the neighborhood
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab about once a month \tab about once a year \tab almost daily \tab don't know \tab never \tab no answer \tab once or twice a week \tab several times a month \tab several times a year \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 327 \tab 121 \tab 35 \tab 5 \tab 164 \tab 1 \tab 290 \tab 271 \tab 270 \tab - \tab 1484 \cr
   1975 \tab - \tab 337 \tab 114 \tab 51 \tab 3 \tab 186 \tab 2 \tab 266 \tab 259 \tab 272 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 338 \tab 110 \tab 43 \tab 3 \tab 153 \tab 4 \tab 287 \tab 301 \tab 291 \tab - \tab 1530 \cr
   1978 \tab - \tab 252 \tab 130 \tab 33 \tab 3 \tab 179 \tab 3 \tab 289 \tab 319 \tab 324 \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 389 \tab 143 \tab 63 \tab 4 \tab 173 \tab 8 \tab 341 \tab 378 \tab 361 \tab - \tab 1860 \cr
   1983 \tab - \tab 354 \tab 98 \tab 41 \tab 2 \tab 169 \tab 3 \tab 301 \tab 339 \tab 292 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 320 \tab 139 \tab 44 \tab 1 \tab 150 \tab 4 \tab 291 \tab 304 \tab 281 \tab - \tab 1534 \cr
   1986 \tab - \tab 331 \tab 79 \tab 24 \tab 1 \tab 171 \tab 4 \tab 299 \tab 269 \tab 292 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 232 \tab 62 \tab 29 \tab - \tab 100 \tab 2 \tab 174 \tab 205 \tab 184 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 229 \tab 73 \tab 40 \tab 2 \tab 115 \tab 2 \tab 184 \tab 180 \tab 181 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 219 \tab 58 \tab 23 \tab 1 \tab 97 \tab 2 \tab 160 \tab 193 \tab 175 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 242 \tab 74 \tab 38 \tab 1 \tab 110 \tab 4 \tab 214 \tab 187 \tab 154 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 214 \tab 65 \tab 35 \tab 1 \tab 97 \tab 3 \tab 227 \tab 220 \tab 218 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 445 \tab 139 \tab 48 \tab 1 \tab 181 \tab 10 \tab 391 \tab 408 \tab 354 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 427 \tab 142 \tab 71 \tab 4 \tab 177 \tab 1 \tab 410 \tab 400 \tab 328 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 383 \tab 126 \tab 64 \tab 10 \tab 158 \tab 5 \tab 347 \tab 399 \tab 379 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 401 \tab 137 \tab 69 \tab 7 \tab 181 \tab 8 \tab 357 \tab 379 \tab 338 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 196 \tab 60 \tab 45 \tab 2 \tab 67 \tab - \tab 178 \tab 221 \tab 139 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 216 \tab 54 \tab 31 \tab 2 \tab 85 \tab - \tab 162 \tab 179 \tab 177 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 462 \tab 155 \tab 64 \tab 1 \tab 205 \tab 2 \tab 390 \tab 370 \tab 343 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 277 \tab 101 \tab 59 \tab 4 \tab 145 \tab - \tab 245 \tab 263 \tab 235 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 308 \tab 90 \tab 41 \tab 3 \tab 135 \tab 2 \tab 289 \tab 311 \tab 251 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 280 \tab 110 \tab 48 \tab 2 \tab 141 \tab 1 \tab 246 \tab 255 \tab 219 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 385 \tab 140 \tab 52 \tab - \tab 153 \tab - \tab 278 \tab 367 \tab 300 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 448 \tab 168 \tab 70 \tab 1 \tab 167 \tab 1 \tab 276 \tab 407 \tab 350 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 378 \tab 131 \tab 32 \tab 1 \tab 157 \tab 1 \tab 244 \tab 338 \tab 277 \tab - \tab 2348 \cr
   2021 \tab 1318 \tab 554 \tab 273 \tab 34 \tab - \tab 343 \tab 1 \tab 253 \tab 467 \tab 777 \tab 12 \tab 4032 \cr
   2022 \tab 1168 \tab 494 \tab 221 \tab 68 \tab 2 \tab 305 \tab - \tab 343 \tab 447 \tab 488 \tab 8 \tab 3544 \cr
   Total \tab 29101 \tab 9438 \tab 3313 \tab 1295 \tab 67 \tab 4464 \tab 74 \tab 7732 \tab 8636 \tab 8250 \tab 20 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} almost daily
\item \code{2} once or twice a week
\item \code{3} several times a month
\item \code{4} about once a month
\item \code{5} several times a year
\item \code{6} about once a year
\item \code{7} never
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
