% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_58.R
\name{socommun}
\alias{socommun}
\title{socommun}
\source{
General Social Survey https://gss.norc.org
}
\description{
Spend evening with neighbor
}
\details{
Question 173. Would you use this card and tell me which answer comes closest to how often you do the following things: b. Spend a social evening with someone who lives in your neighborhood.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab about once a month \tab about once a year \tab almost daily \tab don't know \tab never \tab no answer \tab once or twice a week \tab several times a month \tab several times a year \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 249 \tab 91 \tab 110 \tab 6 \tab 322 \tab 2 \tab 347 \tab 189 \tab 168 \tab - \tab 1484 \cr
   1975 \tab - \tab 218 \tab 117 \tab 89 \tab 3 \tab 350 \tab 2 \tab 334 \tab 196 \tab 181 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 239 \tab 123 \tab 86 \tab 2 \tab 370 \tab 4 \tab 331 \tab 184 \tab 191 \tab - \tab 1530 \cr
   1978 \tab - \tab 189 \tab 130 \tab 92 \tab 6 \tab 379 \tab 4 \tab 359 \tab 167 \tab 206 \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 305 \tab 182 \tab 122 \tab 1 \tab 407 \tab 8 \tab 347 \tab 234 \tab 254 \tab - \tab 1860 \cr
   1983 \tab - \tab 230 \tab 148 \tab 91 \tab 3 \tab 415 \tab 4 \tab 316 \tab 194 \tab 198 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 211 \tab 161 \tab 85 \tab 3 \tab 395 \tab 4 \tab 276 \tab 200 \tab 199 \tab - \tab 1534 \cr
   1986 \tab - \tab 215 \tab 136 \tab 104 \tab - \tab 378 \tab 3 \tab 308 \tab 165 \tab 161 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 149 \tab 69 \tab 46 \tab 2 \tab 279 \tab 2 \tab 203 \tab 110 \tab 128 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 149 \tab 106 \tab 51 \tab 5 \tab 305 \tab 2 \tab 170 \tab 96 \tab 122 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 134 \tab 88 \tab 52 \tab 5 \tab 274 \tab 2 \tab 165 \tab 105 \tab 103 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 140 \tab 78 \tab 55 \tab 2 \tab 311 \tab 3 \tab 191 \tab 119 \tab 125 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 156 \tab 102 \tab 49 \tab 2 \tab 320 \tab 2 \tab 181 \tab 130 \tab 138 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 278 \tab 188 \tab 103 \tab 3 \tab 606 \tab 9 \tab 315 \tab 227 \tab 248 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 307 \tab 182 \tab 101 \tab 6 \tab 561 \tab 2 \tab 330 \tab 223 \tab 248 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 292 \tab 193 \tab 98 \tab 6 \tab 523 \tab 3 \tab 299 \tab 196 \tab 261 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 255 \tab 197 \tab 97 \tab 9 \tab 573 \tab 7 \tab 313 \tab 206 \tab 220 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 108 \tab 94 \tab 82 \tab 3 \tab 272 \tab - \tab 133 \tab 113 \tab 103 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 140 \tab 81 \tab 59 \tab 2 \tab 252 \tab - \tab 138 \tab 109 \tab 125 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 288 \tab 210 \tab 95 \tab - \tab 588 \tab 2 \tab 326 \tab 208 \tab 275 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 183 \tab 143 \tab 83 \tab 2 \tab 398 \tab - \tab 209 \tab 140 \tab 171 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 198 \tab 138 \tab 97 \tab 3 \tab 400 \tab 1 \tab 254 \tab 151 \tab 188 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 202 \tab 127 \tab 67 \tab 3 \tab 428 \tab 1 \tab 191 \tab 148 \tab 135 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 243 \tab 163 \tab 82 \tab - \tab 522 \tab - \tab 272 \tab 189 \tab 204 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 295 \tab 164 \tab 102 \tab 1 \tab 611 \tab 1 \tab 281 \tab 216 \tab 217 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 243 \tab 145 \tab 70 \tab - \tab 510 \tab 2 \tab 219 \tab 174 \tab 196 \tab - \tab 2348 \cr
   2021 \tab 1318 \tab 322 \tab 323 \tab 39 \tab - \tab 1076 \tab - \tab 235 \tab 272 \tab 438 \tab 9 \tab 4032 \cr
   2022 \tab 1168 \tab 303 \tab 226 \tab 83 \tab 2 \tab 864 \tab 1 \tab 320 \tab 264 \tab 312 \tab 1 \tab 3544 \cr
   Total \tab 29101 \tab 6241 \tab 4105 \tab 2290 \tab 80 \tab 12689 \tab 71 \tab 7363 \tab 4925 \tab 5515 \tab 10 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} almost daily
\item \code{2} once or twice a week
\item \code{3} several times a month
\item \code{4} about once a month
\item \code{5} several times a year
\item \code{6} about once a year
\item \code{7} never
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
