% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_58.R
\name{socpars}
\alias{socpars}
\title{socpars}
\source{
General Social Survey https://gss.norc.org
}
\description{
Spend evening with parents
}
\details{
Question 173. Would you use this card and tell me which answer comes closest to how often you do the following things: e. Spend a social evening with your parents
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab almost daily \tab don't know \tab never \tab no answer \tab no such people \tab once a month \tab once a year \tab sev times a mnth \tab sev times a week \tab sev times a year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 49 \tab 2 \tab 109 \tab 4 \tab 551 \tab 113 \tab 118 \tab 166 \tab 216 \tab 204 \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 155 \tab 11 \tab 181 \tab 14 \tab 545 \tab 189 \tab 134 \tab 161 \tab 267 \tab 203 \tab 1860 \cr
   1983 \tab - \tab 89 \tab 1 \tab 113 \tab 4 \tab 494 \tab 161 \tab 129 \tab 157 \tab 248 \tab 203 \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 92 \tab 2 \tab 111 \tab 5 \tab 470 \tab 160 \tab 132 \tab 199 \tab 173 \tab 190 \tab 1534 \cr
   1986 \tab - \tab 100 \tab 2 \tab 106 \tab 7 \tab 443 \tab 150 \tab 128 \tab 113 \tab 220 \tab 201 \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 69 \tab 7 \tab 124 \tab 7 \tab 244 \tab 100 \tab 78 \tab 92 \tab 133 \tab 134 \tab 1481 \cr
   1989 \tab 531 \tab 66 \tab 2 \tab 124 \tab 2 \tab 236 \tab 99 \tab 80 \tab 92 \tab 160 \tab 145 \tab 1537 \cr
   1990 \tab 444 \tab 53 \tab 2 \tab 68 \tab 5 \tab 262 \tab 88 \tab 78 \tab 103 \tab 134 \tab 135 \tab 1372 \cr
   1991 \tab 493 \tab 66 \tab 4 \tab 122 \tab 3 \tab 259 \tab 102 \tab 68 \tab 113 \tab 154 \tab 133 \tab 1517 \cr
   1993 \tab 526 \tab 73 \tab 2 \tab 95 \tab 5 \tab 283 \tab 121 \tab 92 \tab 107 \tab 154 \tab 148 \tab 1606 \cr
   1994 \tab 2485 \tab 27 \tab 2 \tab 75 \tab 1 \tab 127 \tab 47 \tab 51 \tab 46 \tab 66 \tab 65 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 58862 \tab 839 \tab 37 \tab 1228 \tab 57 \tab 3914 \tab 1330 \tab 1088 \tab 1349 \tab 1925 \tab 1761 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} no such people
\item \code{1} almost daily
\item \code{2} sev times a week
\item \code{3} sev times a mnth
\item \code{4} once a month
\item \code{5} sev times a year
\item \code{6} once a year
\item \code{7} never
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
