% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_59.R
\name{socsibs}
\alias{socsibs}
\title{socsibs}
\source{
General Social Survey https://gss.norc.org
}
\description{
Spend evening with siblings
}
\details{
Question 173. Would you use this card and tell me which answer comes closest to how often you do the following things: f. Spend a social evening with a brother or sister
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab almost daily \tab don't know \tab never \tab no answer \tab no such people \tab once a month \tab once a year \tab sev times a mnth \tab sev times a week \tab sev times a year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 71 \tab 4 \tab 153 \tab 5 \tab 168 \tab 167 \tab 245 \tab 182 \tab 208 \tab 329 \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 126 \tab 11 \tab 151 \tab 10 \tab 188 \tab 248 \tab 268 \tab 220 \tab 259 \tab 379 \tab 1860 \cr
   1983 \tab - \tab 70 \tab 5 \tab 163 \tab 3 \tab 169 \tab 204 \tab 207 \tab 208 \tab 221 \tab 349 \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 70 \tab 4 \tab 145 \tab 4 \tab 141 \tab 197 \tab 259 \tab 222 \tab 168 \tab 324 \tab 1534 \cr
   1986 \tab - \tab 88 \tab 5 \tab 148 \tab 7 \tab 114 \tab 208 \tab 229 \tab 162 \tab 190 \tab 319 \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 57 \tab 2 \tab 106 \tab 11 \tab 74 \tab 131 \tab 148 \tab 109 \tab 137 \tab 213 \tab 1481 \cr
   1989 \tab 531 \tab 52 \tab 2 \tab 102 \tab 3 \tab 68 \tab 144 \tab 158 \tab 104 \tab 132 \tab 241 \tab 1537 \cr
   1990 \tab 444 \tab 61 \tab - \tab 84 \tab 5 \tab 86 \tab 134 \tab 131 \tab 91 \tab 121 \tab 215 \tab 1372 \cr
   1991 \tab 493 \tab 50 \tab 6 \tab 95 \tab 4 \tab 82 \tab 150 \tab 157 \tab 118 \tab 126 \tab 236 \tab 1517 \cr
   1993 \tab 526 \tab 52 \tab 3 \tab 112 \tab 9 \tab 95 \tab 152 \tab 158 \tab 140 \tab 128 \tab 231 \tab 1606 \cr
   1994 \tab 2485 \tab 17 \tab 1 \tab 67 \tab 1 \tab 45 \tab 54 \tab 92 \tab 57 \tab 63 \tab 110 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 58862 \tab 714 \tab 43 \tab 1326 \tab 62 \tab 1230 \tab 1789 \tab 2052 \tab 1613 \tab 1753 \tab 2946 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} no such people
\item \code{1} almost daily
\item \code{2} sev times a week
\item \code{3} sev times a mnth
\item \code{4} once a month
\item \code{5} sev times a year
\item \code{6} once a year
\item \code{7} never
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
