% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_71.R
\name{spanking}
\alias{spanking}
\title{Favor spanking to discipline child}
\source{
General Social Survey https://gss.norc.org
}
\description{
spanking
}
\details{
Question 225. Do you strongly agree, agree, disagree, or strongly disagree that it is sometimes necessary to discipline a child with a good, hard, spanking?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab strongly agree \tab strongly disagree \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 816 \tab 195 \tab 9 \tab 1 \tab 404 \tab 45 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 476 \tab 145 \tab 8 \tab 2 \tab 302 \tab 55 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 455 \tab 169 \tab 9 \tab 3 \tab 315 \tab 55 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 466 \tab 154 \tab 8 \tab 3 \tab 261 \tab 36 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 485 \tab 194 \tab 15 \tab 2 \tab 263 \tab 65 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 538 \tab 210 \tab 15 \tab 4 \tab 243 \tab 70 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 909 \tab 357 \tab 20 \tab 7 \tab 522 \tab 162 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 890 \tab 357 \tab 33 \tab 4 \tab 512 \tab 164 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 880 \tab 336 \tab 34 \tab 2 \tab 486 \tab 133 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 780 \tab 328 \tab 33 \tab 4 \tab 588 \tab 144 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 387 \tab 176 \tab 6 \tab 3 \tab 272 \tab 64 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 433 \tab 182 \tab 8 \tab 1 \tab 222 \tab 60 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 893 \tab 402 \tab 20 \tab 2 \tab 532 \tab 143 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 608 \tab 280 \tab 18 \tab 3 \tab 342 \tab 78 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 647 \tab 324 \tab 10 \tab 3 \tab 336 \tab 110 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 598 \tab 277 \tab 21 \tab 4 \tab 302 \tab 100 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 777 \tab 365 \tab 12 \tab 4 \tab 378 \tab 139 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 817 \tab 440 \tab 15 \tab 7 \tab 457 \tab 152 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 700 \tab 368 \tab 23 \tab 7 \tab 340 \tab 121 \tab - \tab 2348 \cr
   2021 \tab 1317 \tab 977 \tab 827 \tab 4 \tab 1 \tab 424 \tab 456 \tab 26 \tab 4032 \cr
   2022 \tab 1168 \tab 887 \tab 634 \tab 23 \tab 2 \tab 434 \tab 389 \tab 7 \tab 3544 \cr
   Total \tab 40129 \tab 14419 \tab 6720 \tab 344 \tab 69 \tab 7935 \tab 2741 \tab 33 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
