% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_40.R
\name{spden}
\alias{spden}
\title{Specific denomination, spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
spden
}
\details{
Question 116. What is your (SPOUSE)'s religious preference? Is it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllll}{
   year \tab iap \tab baptist, don't know which \tab episcopal church \tab lutheran, don't know which \tab methodist, don't know which \tab no answer \tab no denomination or non-denominational church \tab other, specify \tab presbyterian, don't know which \tab don't know \tab african methodist episcopal church \tab american baptist association \tab american baptist churches in the u.s.a. \tab american lutheran church \tab lutheran church - missouri synod \tab lutheran church in america \tab national baptist convention of america \tab national baptist convention u.s.a., inc. \tab other baptist churches \tab other lutheran churches \tab other methodist churches \tab other presbyterian churches \tab presbyterian church in the u.s. \tab southern baptist convention \tab united methodist church \tab united presbyterian church in the u.s.a. \tab wisconsin evangelical lutheran synod \tab african methodist episcopal zion church \tab evangelical lutheran \tab presbyterian, merged \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 796 \tab 207 \tab 33 \tab 79 \tab 165 \tab 1 \tab 32 \tab 142 \tab 49 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 779 \tab 219 \tab 27 \tab 94 \tab 148 \tab 4 \tab 31 \tab 120 \tab 62 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 814 \tab 214 \tab 28 \tab 103 \tab 108 \tab 1 \tab 45 \tab 117 \tab 59 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 864 \tab 190 \tab 29 \tab 82 \tab 123 \tab 4 \tab 33 \tab 125 \tab 47 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 888 \tab 206 \tab 20 \tab 83 \tab 124 \tab 3 \tab 37 \tab 117 \tab 52 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 914 \tab 190 \tab 20 \tab 81 \tab 110 \tab 4 \tab 41 \tab 126 \tab 45 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 897 \tab 177 \tab 24 \tab 51 \tab 105 \tab 4 \tab 25 \tab 125 \tab 53 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1193 \tab 236 \tab 25 \tab 78 \tab 113 \tab 8 \tab 32 \tab 134 \tab 41 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 993 \tab 168 \tab 22 \tab 89 \tab 89 \tab 6 \tab 32 \tab 149 \tab 51 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 926 \tab 37 \tab 15 \tab 23 \tab 9 \tab 11 \tab 35 \tab 141 \tab 3 \tab - \tab 1 \tab 11 \tab 10 \tab 18 \tab 17 \tab 6 \tab 1 \tab 4 \tab 29 \tab 3 \tab 7 \tab 3 \tab 15 \tab 55 \tab 71 \tab 21 \tab 1 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 970 \tab 38 \tab 24 \tab 15 \tab 11 \tab 5 \tab 30 \tab 124 \tab 12 \tab - \tab 5 \tab 15 \tab 1 \tab 27 \tab 21 \tab 2 \tab 2 \tab 2 \tab 35 \tab 5 \tab 5 \tab 5 \tab 12 \tab 80 \tab 73 \tab 12 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 953 \tab 35 \tab 23 \tab 7 \tab 14 \tab 4 \tab 35 \tab 116 \tab 12 \tab 1 \tab 3 \tab 7 \tab 5 \tab 19 \tab 25 \tab 5 \tab 5 \tab 1 \tab 32 \tab 6 \tab 4 \tab 3 \tab 10 \tab 68 \tab 62 \tab 12 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1185 \tab 59 \tab 14 \tab 11 \tab 23 \tab 11 \tab 31 \tab 149 \tab 14 \tab - \tab 6 \tab 12 \tab 9 \tab 21 \tab 10 \tab 10 \tab 12 \tab 4 \tab 19 \tab 1 \tab 5 \tab 4 \tab 7 \tab 110 \tab 70 \tab 18 \tab - \tab 4 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 966 \tab 29 \tab 16 \tab 15 \tab 3 \tab - \tab 35 \tab 129 \tab 10 \tab - \tab 2 \tab 10 \tab 6 \tab 7 \tab 10 \tab 9 \tab 2 \tab 2 \tab 28 \tab 4 \tab 6 \tab 3 \tab 8 \tab 94 \tab 67 \tab 14 \tab 3 \tab - \tab 3 \tab - \tab - \tab 1481 \cr
   1989 \tab 1013 \tab 49 \tab 21 \tab 7 \tab 7 \tab 10 \tab 52 \tab 107 \tab 3 \tab - \tab 2 \tab 7 \tab 4 \tab 17 \tab 18 \tab 6 \tab 6 \tab 1 \tab 17 \tab 3 \tab 2 \tab 5 \tab 11 \tab 76 \tab 68 \tab 9 \tab 2 \tab 2 \tab 8 \tab 4 \tab - \tab 1537 \cr
   1990 \tab 925 \tab 26 \tab 16 \tab 11 \tab 2 \tab 5 \tab 43 \tab 120 \tab 8 \tab 1 \tab 3 \tab 6 \tab - \tab 22 \tab 9 \tab 2 \tab 2 \tab 1 \tab 8 \tab 4 \tab 1 \tab 2 \tab 7 \tab 63 \tab 44 \tab 11 \tab 5 \tab 2 \tab 11 \tab 12 \tab - \tab 1372 \cr
   1991 \tab 1016 \tab 43 \tab 19 \tab 10 \tab 8 \tab 4 \tab 27 \tab 137 \tab 7 \tab - \tab 1 \tab 22 \tab 3 \tab 14 \tab 7 \tab 6 \tab 2 \tab 2 \tab 13 \tab 1 \tab 2 \tab 4 \tab 9 \tab 65 \tab 64 \tab 16 \tab - \tab 1 \tab 9 \tab 5 \tab - \tab 1517 \cr
   1993 \tab 1061 \tab 49 \tab 18 \tab 9 \tab 16 \tab 9 \tab 46 \tab 123 \tab 5 \tab 1 \tab 4 \tab 12 \tab 6 \tab 11 \tab 24 \tab 7 \tab 2 \tab 2 \tab 14 \tab 4 \tab 2 \tab 5 \tab 12 \tab 70 \tab 70 \tab 6 \tab 2 \tab 3 \tab 10 \tab 3 \tab - \tab 1606 \cr
   1994 \tab 2506 \tab 38 \tab 22 \tab 9 \tab 17 \tab 23 \tab 34 \tab 87 \tab 2 \tab 3 \tab 5 \tab 14 \tab 2 \tab 17 \tab 12 \tab 6 \tab 1 \tab 3 \tab 16 \tab 3 \tab 5 \tab 4 \tab 6 \tab 68 \tab 55 \tab 15 \tab 2 \tab - \tab 8 \tab 9 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2431 \tab 44 \tab 8 \tab 16 \tab 10 \tab 2 \tab 68 \tab 80 \tab 7 \tab 4 \tab 1 \tab 9 \tab 2 \tab 3 \tab 8 \tab 1 \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab 2 \tab 5 \tab 57 \tab 34 \tab 5 \tab 3 \tab - \tab 4 \tab - \tab - \tab 2812 \cr
   2006 \tab 3765 \tab 76 \tab 23 \tab 19 \tab 17 \tab 2 \tab 95 \tab 195 \tab 24 \tab 11 \tab 1 \tab 10 \tab 7 \tab 4 \tab 17 \tab 5 \tab 1 \tab 5 \tab 18 \tab - \tab 1 \tab 1 \tab 3 \tab 110 \tab 69 \tab 8 \tab 8 \tab 2 \tab 7 \tab 6 \tab - \tab 4510 \cr
   2008 \tab 1461 \tab 63 \tab 27 \tab 17 \tab 12 \tab 10 \tab 95 \tab 140 \tab 21 \tab 3 \tab 2 \tab 10 \tab 3 \tab 4 \tab 10 \tab 2 \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab 67 \tab 48 \tab 4 \tab 5 \tab - \tab 6 \tab - \tab - \tab 2023 \cr
   2010 \tab 1573 \tab 63 \tab 26 \tab 17 \tab 14 \tab 1 \tab 72 \tab 99 \tab 12 \tab 7 \tab - \tab 4 \tab 3 \tab 3 \tab 5 \tab 1 \tab 3 \tab 1 \tab 11 \tab - \tab - \tab 2 \tab 2 \tab 66 \tab 33 \tab 5 \tab 4 \tab - \tab 9 \tab 8 \tab - \tab 2044 \cr
   2012 \tab 1476 \tab 42 \tab 15 \tab 13 \tab 10 \tab 2 \tab 104 \tab 125 \tab 13 \tab 7 \tab 1 \tab 12 \tab 4 \tab 5 \tab 7 \tab 1 \tab - \tab - \tab 8 \tab 3 \tab 2 \tab - \tab 2 \tab 61 \tab 41 \tab 2 \tab 5 \tab - \tab 7 \tab 6 \tab - \tab 1974 \cr
   2014 \tab 1936 \tab 45 \tab 17 \tab 18 \tab 14 \tab 12 \tab 152 \tab 115 \tab 14 \tab 7 \tab 6 \tab 11 \tab 2 \tab 4 \tab 15 \tab 2 \tab 5 \tab - \tab 8 \tab 3 \tab 1 \tab 3 \tab 3 \tab 65 \tab 58 \tab 5 \tab 1 \tab - \tab 13 \tab 3 \tab - \tab 2538 \cr
   2016 \tab 2279 \tab 50 \tab 12 \tab 17 \tab 11 \tab 7 \tab 130 \tab 137 \tab 10 \tab 12 \tab 4 \tab 11 \tab 6 \tab 2 \tab 16 \tab 6 \tab 4 \tab 2 \tab 12 \tab 3 \tab 1 \tab 4 \tab - \tab 72 \tab 39 \tab 1 \tab 5 \tab 1 \tab 9 \tab 4 \tab - \tab 2867 \cr
   2018 \tab 1853 \tab 55 \tab 11 \tab 10 \tab 15 \tab 4 \tab 114 \tab 112 \tab 11 \tab 11 \tab - \tab 7 \tab 4 \tab 3 \tab 7 \tab 2 \tab 1 \tab 1 \tab 11 \tab - \tab 2 \tab 5 \tab 3 \tab 54 \tab 35 \tab 2 \tab 5 \tab 1 \tab 5 \tab 4 \tab - \tab 2348 \cr
   2021 \tab 3181 \tab 99 \tab 27 \tab 39 \tab 20 \tab 3 \tab 250 \tab 123 \tab 30 \tab 15 \tab 3 \tab 12 \tab 8 \tab 2 \tab 20 \tab 4 \tab 4 \tab 5 \tab 6 \tab 6 \tab - \tab 5 \tab 5 \tab 56 \tab 71 \tab 3 \tab 5 \tab 1 \tab 16 \tab 10 \tab 3 \tab 4032 \cr
   2022 \tab 2904 \tab 69 \tab 19 \tab 18 \tab 12 \tab 22 \tab 193 \tab 97 \tab 13 \tab - \tab 1 \tab 14 \tab 7 \tab 4 \tab 15 \tab 3 \tab 2 \tab 2 \tab 12 \tab - \tab 3 \tab 5 \tab 5 \tab 59 \tab 43 \tab 5 \tab 1 \tab - \tab 10 \tab 6 \tab - \tab 3544 \cr
   Total \tab 55449 \tab 2816 \tab 601 \tab 1041 \tab 1330 \tab 182 \tab 1949 \tab 3611 \tab 690 \tab 94 \tab 51 \tab 216 \tab 92 \tab 207 \tab 273 \tab 86 \tab 61 \tab 42 \tab 302 \tab 49 \tab 51 \tab 65 \tab 129 \tab 1416 \tab 1115 \tab 174 \tab 60 \tab 20 \tab 135 \tab 80 \tab 3 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{10} american baptist association
\item \code{11} american baptist churches in the u.s.a.
\item \code{12} national baptist convention of america
\item \code{13} national baptist convention u.s.a., inc.
\item \code{14} southern baptist convention
\item \code{15} other baptist churches
\item \code{18} baptist, don't know which
\item \code{20} african methodist episcopal church
\item \code{21} african methodist episcopal zion church
\item \code{22} united methodist church
\item \code{23} other methodist churches
\item \code{28} methodist, don't know which
\item \code{30} american lutheran church
\item \code{31} lutheran church in america
\item \code{32} lutheran church - missouri synod
\item \code{33} wisconsin evangelical lutheran synod
\item \code{34} other lutheran churches
\item \code{35} evangelical lutheran
\item \code{38} lutheran, don't know which
\item \code{40} presbyterian church in the u.s.
\item \code{41} united presbyterian church in the u.s.a.
\item \code{42} other presbyterian churches
\item \code{43} presbyterian, merged
\item \code{48} presbyterian, don't know which
\item \code{50} episcopal church
\item \code{60} other, specify
\item \code{70} no denomination or non-denominational church
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
