% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_41.R
\name{spden16}
\alias{spden16}
\title{spden16}
\source{
General Social Survey https://gss.norc.org
}
\description{
Denomination in which spouse raised
}
\details{
Question 117a. If Protestant: What specific denomination is that, if any?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllll}{
   year \tab iap \tab baptist-dk which \tab episcopal \tab lutheran-dk which \tab methodist-dk which \tab no answer \tab no denomination \tab other \tab presbyterian-dk wh \tab don't know \tab afr meth episcopal \tab am bapt ch in usa \tab am baptist asso \tab am lutheran \tab luth ch in america \tab lutheran-mo synod \tab nat bapt conv of am \tab nat bapt conv usa \tab other baptists \tab other lutheran \tab other methodist \tab other presbyterian \tab presbyterian c in us \tab southern baptist \tab united methodist \tab united pres ch in usa \tab afr meth ep zion \tab wi evan luth synod \tab evangelic luth \tab presbyterian, merged \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 772 \tab 240 \tab 22 \tab 76 \tab 173 \tab 2 \tab 40 \tab 132 \tab 47 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 776 \tab 236 \tab 26 \tab 101 \tab 156 \tab 2 \tab 24 \tab 113 \tab 48 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 804 \tab 234 \tab 27 \tab 101 \tab 137 \tab - \tab 27 \tab 105 \tab 55 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 849 \tab 204 \tab 28 \tab 95 \tab 141 \tab 2 \tab 17 \tab 113 \tab 45 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 877 \tab 232 \tab 29 \tab 82 \tab 129 \tab 6 \tab 20 \tab 99 \tab 54 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 883 \tab 205 \tab 25 \tab 75 \tab 130 \tab 6 \tab 36 \tab 125 \tab 46 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 889 \tab 181 \tab 18 \tab 56 \tab 119 \tab 5 \tab 15 \tab 118 \tab 50 \tab 17 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1184 \tab 263 \tab 25 \tab 69 \tab 126 \tab 10 \tab 27 \tab 115 \tab 41 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 987 \tab 185 \tab 23 \tab 83 \tab 112 \tab 10 \tab 18 \tab 121 \tab 55 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 941 \tab 52 \tab 16 \tab 25 \tab 20 \tab 12 \tab 25 \tab 117 \tab 10 \tab - \tab 1 \tab 10 \tab 13 \tab 16 \tab 5 \tab 14 \tab 2 \tab 4 \tab 24 \tab 6 \tab 8 \tab 2 \tab 14 \tab 70 \tab 53 \tab 13 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 972 \tab 47 \tab 24 \tab 20 \tab 24 \tab 8 \tab 24 \tab 104 \tab 14 \tab - \tab 6 \tab 3 \tab 18 \tab 24 \tab 1 \tab 19 \tab 1 \tab 1 \tab 30 \tab 4 \tab 14 \tab 4 \tab 12 \tab 78 \tab 70 \tab 9 \tab 1 \tab 2 \tab - \tab - \tab 1534 \cr
   1986 \tab 968 \tab 45 \tab 23 \tab 12 \tab 15 \tab 10 \tab 14 \tab 91 \tab 11 \tab 4 \tab 3 \tab 5 \tab 9 \tab 17 \tab 4 \tab 19 \tab 6 \tab 1 \tab 32 \tab 4 \tab 5 \tab 3 \tab 7 \tab 79 \tab 67 \tab 13 \tab 1 \tab 2 \tab - \tab - \tab 1470 \cr
   1987 \tab 1190 \tab 73 \tab 17 \tab 16 \tab 33 \tab 12 \tab 24 \tab 119 \tab 16 \tab - \tab 7 \tab 10 \tab 11 \tab 19 \tab 4 \tab 8 \tab 12 \tab 6 \tab 16 \tab 6 \tab 5 \tab 4 \tab 4 \tab 107 \tab 80 \tab 18 \tab 2 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 958 \tab 42 \tab 15 \tab 23 \tab 15 \tab 2 \tab 28 \tab 112 \tab 18 \tab - \tab 6 \tab 5 \tab 11 \tab 15 \tab 7 \tab 9 \tab 2 \tab 2 \tab 20 \tab 6 \tab 4 \tab 1 \tab 5 \tab 87 \tab 68 \tab 18 \tab - \tab 2 \tab - \tab - \tab 1481 \cr
   1989 \tab 1012 \tab 61 \tab 19 \tab 15 \tab 19 \tab 11 \tab 32 \tab 106 \tab 5 \tab - \tab 2 \tab 3 \tab 11 \tab 18 \tab 5 \tab 14 \tab 6 \tab 1 \tab 13 \tab 6 \tab 6 \tab 2 \tab 9 \tab 76 \tab 65 \tab 11 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1537 \cr
   1990 \tab 926 \tab 27 \tab 16 \tab 15 \tab 13 \tab 7 \tab 31 \tab 97 \tab 10 \tab 5 \tab 5 \tab - \tab 7 \tab 24 \tab 2 \tab 7 \tab 1 \tab 1 \tab 5 \tab 5 \tab 3 \tab 1 \tab 8 \tab 66 \tab 53 \tab 12 \tab 2 \tab 6 \tab 8 \tab 9 \tab 1372 \cr
   1991 \tab 1015 \tab 45 \tab 16 \tab 17 \tab 22 \tab 9 \tab 23 \tab 116 \tab 10 \tab 1 \tab 2 \tab 3 \tab 28 \tab 20 \tab 2 \tab 9 \tab 3 \tab 1 \tab 6 \tab 2 \tab 3 \tab 1 \tab 8 \tab 78 \tab 60 \tab 9 \tab 1 \tab - \tab 5 \tab 2 \tab 1517 \cr
   1993 \tab 1054 \tab 61 \tab 18 \tab 11 \tab 21 \tab 14 \tab 36 \tab 110 \tab 5 \tab 4 \tab 3 \tab 6 \tab 15 \tab 17 \tab 5 \tab 15 \tab 3 \tab 3 \tab 8 \tab 6 \tab 1 \tab 5 \tab 10 \tab 84 \tab 69 \tab 11 \tab 3 \tab 2 \tab 4 \tab 2 \tab 1606 \cr
   1994 \tab 2525 \tab 55 \tab 18 \tab 9 \tab 15 \tab 28 \tab 19 \tab 52 \tab - \tab 2 \tab 7 \tab 2 \tab 16 \tab 20 \tab 3 \tab 12 \tab 1 \tab 2 \tab 15 \tab 4 \tab 3 \tab 3 \tab 10 \tab 78 \tab 59 \tab 11 \tab - \tab 3 \tab 8 \tab 12 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 61205 \tab 2488 \tab 405 \tab 901 \tab 1420 \tab 156 \tab 480 \tab 2065 \tab 540 \tab 48 \tab 42 \tab 47 \tab 139 \tab 190 \tab 38 \tab 126 \tab 37 \tab 22 \tab 169 \tab 49 \tab 52 \tab 26 \tab 87 \tab 803 \tab 644 \tab 125 \tab 13 \tab 19 \tab 27 \tab 27 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{10} am baptist asso
\item \code{11} am bapt ch in usa
\item \code{12} nat bapt conv of am
\item \code{13} nat bapt conv usa
\item \code{14} southern baptist
\item \code{15} other baptists
\item \code{18} baptist-dk which
\item \code{20} afr meth episcopal
\item \code{21} afr meth ep zion
\item \code{22} united methodist
\item \code{23} other methodist
\item \code{28} methodist-dk which
\item \code{30} am lutheran
\item \code{31} luth ch in america
\item \code{32} lutheran-mo synod
\item \code{33} wi evan luth synod
\item \code{34} other lutheran
\item \code{35} evangelic luth
\item \code{38} lutheran-dk which
\item \code{40} presbyterian c in us
\item \code{41} united pres ch in usa
\item \code{42} other presbyterian
\item \code{43} presbyterian, merged
\item \code{48} presbyterian-dk wh
\item \code{50} episcopal
\item \code{60} other
\item \code{70} no denomination
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
