% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_18.R
\name{spdotpeo}
\alias{spdotpeo}
\title{Sp. d.o.t. code - relation to people}
\source{
General Social Survey https://gss.norc.org
}
\description{
spdotpeo
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 0 \tab 0.06 \tab 0.54 \tab 0.62 \tab 0.73 \tab 0.9 \tab 1.43 \tab 1.68 \tab 1.89 \tab 1.97 \tab 2 \tab 2.02 \tab 2.03 \tab 2.2 \tab 2.29 \tab 2.35 \tab 2.51 \tab 2.75 \tab 3.33 \tab 3.39 \tab 3.55 \tab 3.68 \tab 3.75 \tab 3.77 \tab 3.83 \tab 3.94 \tab 3.95 \tab 3.96 \tab 3.97 \tab 4 \tab 4.02 \tab 4.28 \tab 4.33 \tab 4.39 \tab 4.4 \tab 4.54 \tab 4.72 \tab 4.99 \tab 5.01 \tab 5.02 \tab 5.06 \tab 5.07 \tab 5.09 \tab 5.13 \tab 5.14 \tab 5.33 \tab 5.36 \tab 5.38 \tab 5.42 \tab 5.43 \tab 5.48 \tab 5.59 \tab 5.62 \tab 5.65 \tab 5.69 \tab 5.74 \tab 5.78 \tab 5.8 \tab 5.81 \tab 5.84 \tab 5.86 \tab 5.88 \tab 5.94 \tab 5.95 \tab 5.99 \tab 6 \tab 6.02 \tab 6.03 \tab 6.04 \tab 6.05 \tab 6.06 \tab 6.07 \tab 6.08 \tab 6.15 \tab 6.17 \tab 6.22 \tab 6.28 \tab 6.29 \tab 6.32 \tab 6.34 \tab 6.41 \tab 6.44 \tab 6.45 \tab 6.46 \tab 6.47 \tab 6.49 \tab 6.5 \tab 6.51 \tab 6.65 \tab 6.79 \tab 6.8 \tab 6.82 \tab 6.88 \tab 6.91 \tab 6.93 \tab 6.94 \tab 6.98 \tab 6.99 \tab 7 \tab 7.01 \tab 7.03 \tab 7.16 \tab 7.21 \tab 7.23 \tab 7.25 \tab 7.27 \tab 7.4 \tab 7.43 \tab 7.45 \tab 7.47 \tab 7.49 \tab 7.51 \tab 7.54 \tab 7.55 \tab 7.57 \tab 7.59 \tab 7.6 \tab 7.62 \tab 7.63 \tab 7.64 \tab 7.65 \tab 7.66 \tab 7.67 \tab 7.68 \tab 7.69 \tab 7.71 \tab 7.74 \tab 7.76 \tab 7.77 \tab 7.78 \tab 7.79 \tab 7.8 \tab 7.81 \tab 7.84 \tab 7.85 \tab 7.86 \tab 7.87 \tab 7.88 \tab 7.89 \tab 7.9 \tab 7.91 \tab 7.92 \tab 7.93 \tab 7.94 \tab 7.95 \tab 7.96 \tab 7.97 \tab 7.98 \tab 8 \tab iap \tab 2.44 \tab 2.59 \tab 3 \tab 3.42 \tab 3.51 \tab 3.85 \tab 4.23 \tab 4.31 \tab 4.84 \tab 4.94 \tab 5.17 \tab 5.49 \tab 5.58 \tab 5.97 \tab 6.25 \tab 6.55 \tab 7.04 \tab 7.08 \tab 7.22 \tab 7.5 \tab 7.58 \tab 7.7 \tab 7.72 \tab 7.73 \tab 7.75 \tab 7.82 \tab 0.42 \tab 1.44 \tab 2.09 \tab 2.49 \tab 3.06 \tab 4.04 \tab 4.15 \tab 4.35 \tab 4.36 \tab 4.7 \tab 5.18 \tab 5.45 \tab 5.5 \tab 5.57 \tab 5.75 \tab 5.9 \tab 6.09 \tab 6.16 \tab 6.62 \tab 7.28 \tab 2.3 \tab 2.47 \tab 3.66 \tab 4.08 \tab 5.26 \tab 7.83 \tab 2.64 \tab 5 \tab 5.37 \tab 6.66 \tab 3.24 \tab 7.05 \tab 7.36 \tab 4.86 \tab 5.28 \tab 6.37 \tab 7.02 \tab 4.73 \tab 4.93 \tab 5.63 \tab 7.12 \tab 2.39 \tab 2.62 \tab 4.1 \tab 4.17 \tab 5.67 \tab 7.07 \tab 7.24 \tab 3.48 \tab 7.14 \tab 7.15 \tab 5.23 \tab 5.27 \tab 5.29 \tab 5.66 \tab 7.11 \tab Total \cr
   1972 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab 10 \tab 25 \tab 5 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 9 \tab 1 \tab 4 \tab 3 \tab 1 \tab 18 \tab 3 \tab 2 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 5 \tab 2 \tab 3 \tab 4 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 6 \tab 5 \tab 24 \tab 32 \tab 7 \tab 1 \tab 2 \tab 2 \tab 2 \tab 5 \tab 2 \tab 6 \tab 16 \tab 6 \tab 9 \tab 5 \tab 1 \tab 3 \tab 67 \tab 2 \tab 3 \tab 9 \tab 9 \tab 2 \tab 3 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab 5 \tab 11 \tab 3 \tab 2 \tab 3 \tab 1 \tab 8 \tab 5 \tab 6 \tab 3 \tab 1 \tab 17 \tab 1 \tab 9 \tab 1 \tab 1 \tab 1 \tab 16 \tab 10 \tab 1 \tab 2 \tab 1 \tab 11 \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab 2 \tab 4 \tab 7 \tab 2 \tab 9 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 11 \tab 8 \tab 4 \tab 2 \tab 16 \tab 14 \tab 3 \tab 4 \tab 8 \tab 7 \tab 7 \tab 5 \tab 6 \tab 6 \tab 24 \tab 23 \tab 31 \tab 2 \tab 2 \tab 11 \tab 9 \tab 27 \tab 6 \tab 11 \tab 10 \tab 27 \tab 8 \tab 62 \tab 18 \tab 90 \tab 594 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 3 \tab 8 \tab 4 \tab 6 \tab 1 \tab 8 \tab 1 \tab 1 \tab 2 \tab 6 \tab 9 \tab 19 \tab 12 \tab 1 \tab 3 \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab 3 \tab 3 \tab 3 \tab 1 \tab 18 \tab 4 \tab - \tab 2 \tab 1 \tab 6 \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 33 \tab 24 \tab 12 \tab - \tab - \tab 3 \tab - \tab 3 \tab 3 \tab 4 \tab 21 \tab 2 \tab 7 \tab 6 \tab 1 \tab 3 \tab 36 \tab - \tab 8 \tab 12 \tab 6 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 15 \tab 7 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 4 \tab - \tab 1 \tab 8 \tab 2 \tab 9 \tab 4 \tab 1 \tab 1 \tab 13 \tab 9 \tab 2 \tab 4 \tab 1 \tab 5 \tab 2 \tab 1 \tab 4 \tab 6 \tab 2 \tab 5 \tab 2 \tab 5 \tab 1 \tab 4 \tab - \tab - \tab - \tab 7 \tab 1 \tab 3 \tab 7 \tab 2 \tab 3 \tab 12 \tab 16 \tab 15 \tab 8 \tab 5 \tab 2 \tab 8 \tab 6 \tab 5 \tab 12 \tab 5 \tab 15 \tab 14 \tab 22 \tab 3 \tab 1 \tab 11 \tab 10 \tab 22 \tab 12 \tab 6 \tab 6 \tab 27 \tab 12 \tab 44 \tab 38 \tab 84 \tab 563 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 3 \tab 2 \tab 4 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab 3 \tab - \tab 17 \tab 21 \tab 2 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 3 \tab 5 \tab 2 \tab 4 \tab 1 \tab - \tab 23 \tab 7 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 4 \tab - \tab 3 \tab 1 \tab 4 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 30 \tab 4 \tab 3 \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 2 \tab 8 \tab 30 \tab 3 \tab 5 \tab 3 \tab - \tab 2 \tab 49 \tab - \tab 4 \tab 13 \tab 8 \tab - \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 19 \tab 4 \tab - \tab 1 \tab 1 \tab 5 \tab 3 \tab 1 \tab 4 \tab 1 \tab 9 \tab 2 \tab 8 \tab 2 \tab - \tab - \tab 16 \tab 8 \tab - \tab 2 \tab 2 \tab 10 \tab 1 \tab 1 \tab 4 \tab 4 \tab 1 \tab - \tab 7 \tab 4 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 9 \tab 1 \tab 4 \tab 6 \tab 4 \tab 1 \tab 7 \tab 13 \tab 13 \tab 11 \tab 2 \tab 2 \tab 8 \tab 5 \tab 3 \tab 10 \tab 6 \tab 22 \tab 10 \tab 29 \tab 1 \tab 1 \tab 12 \tab 10 \tab 26 \tab 10 \tab 17 \tab 24 \tab 15 \tab 7 \tab 45 \tab 34 \tab 78 \tab 540 \tab 1 \tab 1 \tab 1 \tab - \tab 9 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 1 \tab 2 \tab 7 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 15 \tab 23 \tab 3 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 2 \tab 25 \tab 6 \tab - \tab 5 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 5 \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 5 \tab 5 \tab 25 \tab 36 \tab 9 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 3 \tab 4 \tab 8 \tab 2 \tab 5 \tab 1 \tab 1 \tab 4 \tab 37 \tab 1 \tab 2 \tab 15 \tab 12 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 11 \tab 9 \tab 3 \tab 1 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 11 \tab 3 \tab 16 \tab 1 \tab - \tab - \tab 11 \tab 6 \tab - \tab 3 \tab - \tab 9 \tab 1 \tab - \tab 4 \tab 1 \tab 3 \tab - \tab 3 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab 6 \tab - \tab 3 \tab 6 \tab 4 \tab 4 \tab 7 \tab 9 \tab 12 \tab 5 \tab 1 \tab 10 \tab 8 \tab 2 \tab 1 \tab 9 \tab 4 \tab 29 \tab 12 \tab 25 \tab 4 \tab 1 \tab 7 \tab 9 \tab 32 \tab 2 \tab 9 \tab 24 \tab 15 \tab 10 \tab 47 \tab 32 \tab 87 \tab 585 \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1 \tab 5 \tab 3 \tab 3 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab 17 \tab 14 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 6 \tab 2 \tab 5 \tab 4 \tab 1 \tab 18 \tab 2 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 6 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 14 \tab 34 \tab 5 \tab - \tab - \tab 4 \tab - \tab 2 \tab 1 \tab 2 \tab 5 \tab 5 \tab 11 \tab 3 \tab - \tab 10 \tab 37 \tab - \tab 1 \tab 11 \tab 6 \tab 1 \tab 2 \tab 3 \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 12 \tab 7 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 4 \tab 7 \tab 1 \tab 7 \tab - \tab 1 \tab 3 \tab 13 \tab 6 \tab 2 \tab 3 \tab - \tab 3 \tab 3 \tab 1 \tab 1 \tab 6 \tab - \tab 1 \tab 5 \tab 2 \tab 2 \tab 6 \tab - \tab - \tab - \tab 10 \tab 1 \tab 3 \tab - \tab 5 \tab 4 \tab 7 \tab 10 \tab 14 \tab 8 \tab 4 \tab 5 \tab 8 \tab 4 \tab 1 \tab 4 \tab 6 \tab 23 \tab 17 \tab 27 \tab 3 \tab - \tab 12 \tab 10 \tab 26 \tab 10 \tab 16 \tab 8 \tab 15 \tab 10 \tab 52 \tab 25 \tab 86 \tab 633 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab 5 \tab 4 \tab 12 \tab 19 \tab 2 \tab - \tab - \tab - \tab - \tab 6 \tab - \tab - \tab 5 \tab - \tab 6 \tab 1 \tab 1 \tab 29 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 24 \tab 32 \tab 4 \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab 3 \tab 3 \tab 7 \tab 2 \tab 5 \tab 4 \tab 1 \tab 4 \tab 43 \tab - \tab 2 \tab 9 \tab 9 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 12 \tab 9 \tab - \tab 1 \tab 1 \tab 4 \tab 6 \tab 2 \tab 2 \tab 4 \tab 2 \tab 3 \tab 12 \tab - \tab 1 \tab 1 \tab 12 \tab 6 \tab 5 \tab 3 \tab - \tab 7 \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 7 \tab - \tab 1 \tab 1 \tab - \tab - \tab 10 \tab - \tab 3 \tab 13 \tab 3 \tab - \tab 3 \tab 6 \tab 10 \tab 4 \tab 2 \tab 4 \tab 6 \tab 13 \tab 2 \tab 3 \tab 8 \tab 14 \tab 12 \tab 22 \tab 1 \tab - \tab 13 \tab 6 \tab 30 \tab 9 \tab 16 \tab 13 \tab 34 \tab 9 \tab 44 \tab 34 \tab 91 \tab 652 \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 1 \tab 5 \tab 2 \tab 5 \tab 1 \tab 4 \tab - \tab 1 \tab 8 \tab - \tab 3 \tab 22 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 10 \tab - \tab - \tab 8 \tab 1 \tab 2 \tab 6 \tab 1 \tab 24 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab 4 \tab - \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 29 \tab 50 \tab 8 \tab - \tab - \tab - \tab - \tab 5 \tab 2 \tab 3 \tab 2 \tab 2 \tab 13 \tab 3 \tab 1 \tab 1 \tab 48 \tab 1 \tab 5 \tab 8 \tab 1 \tab - \tab 3 \tab - \tab 2 \tab 4 \tab - \tab 3 \tab 1 \tab - \tab 13 \tab 8 \tab 3 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 8 \tab - \tab 11 \tab 1 \tab - \tab 2 \tab 10 \tab 6 \tab 1 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 5 \tab 3 \tab 3 \tab 5 \tab 10 \tab 7 \tab 3 \tab 3 \tab 5 \tab 6 \tab 6 \tab 4 \tab 7 \tab 9 \tab 19 \tab 17 \tab 24 \tab 5 \tab - \tab 10 \tab 4 \tab 15 \tab 12 \tab 9 \tab 14 \tab 12 \tab 10 \tab 39 \tab 27 \tab 85 \tab 682 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 5 \tab 5 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 4 \tab 4 \tab - \tab 4 \tab - \tab 5 \tab - \tab 2 \tab 2 \tab 1 \tab 8 \tab 19 \tab 3 \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 4 \tab 1 \tab 3 \tab 6 \tab 3 \tab 13 \tab 10 \tab - \tab 3 \tab 4 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab 15 \tab 38 \tab 6 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 8 \tab 6 \tab 3 \tab 3 \tab 11 \tab 4 \tab 1 \tab 1 \tab 33 \tab - \tab 3 \tab 1 \tab 12 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 9 \tab 4 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 4 \tab 4 \tab 1 \tab 11 \tab 4 \tab 7 \tab - \tab - \tab 3 \tab 7 \tab 9 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 5 \tab 3 \tab 3 \tab 5 \tab - \tab - \tab - \tab 6 \tab - \tab 5 \tab 6 \tab 12 \tab 1 \tab 6 \tab 10 \tab 12 \tab 6 \tab 2 \tab 4 \tab 4 \tab 7 \tab 3 \tab 4 \tab 7 \tab 14 \tab 15 \tab 27 \tab 3 \tab 2 \tab 13 \tab 5 \tab 20 \tab 6 \tab 7 \tab 5 \tab 18 \tab 1 \tab 39 \tab 15 \tab 79 \tab 653 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 2 \tab 1 \tab 3 \tab 3 \tab 2 \tab 2 \tab 5 \tab 2 \tab 1 \tab 2 \tab 3 \tab 9 \tab 25 \tab 3 \tab - \tab 1 \tab - \tab - \tab 4 \tab 2 \tab - \tab 11 \tab - \tab 1 \tab 4 \tab 2 \tab 10 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 6 \tab 36 \tab 9 \tab - \tab - \tab 2 \tab 2 \tab 5 \tab 6 \tab 7 \tab 3 \tab 1 \tab 7 \tab 5 \tab - \tab 7 \tab 35 \tab 4 \tab 3 \tab 15 \tab 4 \tab - \tab - \tab 3 \tab - \tab 4 \tab - \tab - \tab - \tab 1 \tab 10 \tab 7 \tab 1 \tab 3 \tab 1 \tab 7 \tab 3 \tab 1 \tab 6 \tab 7 \tab 12 \tab 6 \tab 14 \tab 1 \tab 1 \tab 4 \tab 9 \tab 4 \tab 1 \tab 1 \tab - \tab 7 \tab 2 \tab - \tab 2 \tab 1 \tab 5 \tab 1 \tab 4 \tab 6 \tab 1 \tab 12 \tab - \tab - \tab - \tab 6 \tab - \tab - \tab 9 \tab 9 \tab - \tab 6 \tab 9 \tab 13 \tab 11 \tab 3 \tab 3 \tab 9 \tab 5 \tab 6 \tab 5 \tab 7 \tab 15 \tab 16 \tab 24 \tab 3 \tab 1 \tab 14 \tab 8 \tab 20 \tab 3 \tab 9 \tab 16 \tab 19 \tab 15 \tab 41 \tab 33 \tab 81 \tab 957 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 3 \tab 4 \tab 2 \tab 4 \tab 5 \tab 1 \tab 2 \tab - \tab 1 \tab 4 \tab 2 \tab 9 \tab 12 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 10 \tab - \tab - \tab 6 \tab - \tab 4 \tab - \tab - \tab 16 \tab 4 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 1 \tab 5 \tab - \tab - \tab 8 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 7 \tab 48 \tab 9 \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab 6 \tab 6 \tab 4 \tab 3 \tab 8 \tab 5 \tab 2 \tab 6 \tab 49 \tab 1 \tab 2 \tab 12 \tab 13 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 19 \tab 6 \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 3 \tab 2 \tab 9 \tab 4 \tab 14 \tab 1 \tab - \tab 3 \tab 11 \tab 10 \tab - \tab 3 \tab 1 \tab 3 \tab 4 \tab 1 \tab 3 \tab 4 \tab 2 \tab 4 \tab 4 \tab 3 \tab 4 \tab 3 \tab - \tab - \tab - \tab 6 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 7 \tab 5 \tab 16 \tab 10 \tab 1 \tab 4 \tab 7 \tab 5 \tab 2 \tab 13 \tab 5 \tab 16 \tab 11 \tab 19 \tab 3 \tab 2 \tab 8 \tab 5 \tab 27 \tab 7 \tab 4 \tab 9 \tab 23 \tab 9 \tab 42 \tab 19 \tab 73 \tab 711 \tab - \tab 2 \tab - \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab 4 \tab - \tab 5 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1 \tab 2 \tab 5 \tab - \tab 3 \tab - \tab 3 \tab - \tab 1 \tab 3 \tab 2 \tab 12 \tab 11 \tab 2 \tab 1 \tab - \tab - \tab - \tab 4 \tab 1 \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab - \tab 20 \tab - \tab 1 \tab 4 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 12 \tab 29 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 7 \tab 3 \tab 6 \tab 4 \tab 10 \tab 6 \tab - \tab 6 \tab 24 \tab 2 \tab 3 \tab 9 \tab 8 \tab 1 \tab 4 \tab 8 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab 11 \tab 6 \tab - \tab 4 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 14 \tab 1 \tab 12 \tab - \tab - \tab - \tab 10 \tab 11 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 4 \tab - \tab 8 \tab 2 \tab 1 \tab 2 \tab 13 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab 4 \tab 5 \tab 3 \tab 4 \tab 10 \tab 10 \tab 12 \tab - \tab 7 \tab 1 \tab 5 \tab 1 \tab 20 \tab 5 \tab 8 \tab 12 \tab 21 \tab 6 \tab 1 \tab 11 \tab 4 \tab 19 \tab 5 \tab 1 \tab 8 \tab 12 \tab 4 \tab 33 \tab 21 \tab 78 \tab 706 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 4 \tab - \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 3 \tab 2 \tab 3 \tab 3 \tab - \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab 12 \tab 13 \tab 5 \tab - \tab - \tab - \tab - \tab 11 \tab - \tab - \tab 3 \tab - \tab 4 \tab 3 \tab 2 \tab 15 \tab 5 \tab - \tab 2 \tab 5 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab 3 \tab 1 \tab 10 \tab 38 \tab 6 \tab 1 \tab - \tab 3 \tab 1 \tab 4 \tab 10 \tab 2 \tab 4 \tab 2 \tab 14 \tab 9 \tab 1 \tab 5 \tab 27 \tab 1 \tab 3 \tab 10 \tab 7 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 14 \tab 2 \tab 2 \tab 1 \tab - \tab 5 \tab 6 \tab 1 \tab 5 \tab 3 \tab 10 \tab 2 \tab 11 \tab 2 \tab - \tab - \tab 12 \tab 8 \tab 3 \tab 3 \tab - \tab 5 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 5 \tab 1 \tab 2 \tab 8 \tab 6 \tab 2 \tab 6 \tab 12 \tab 13 \tab 7 \tab 3 \tab 5 \tab 8 \tab 5 \tab 1 \tab 14 \tab 8 \tab 9 \tab 9 \tab 17 \tab 7 \tab 1 \tab 9 \tab 4 \tab 33 \tab 6 \tab 9 \tab 9 \tab 22 \tab 10 \tab 37 \tab 28 \tab 64 \tab 714 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 4 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 4 \tab 7 \tab 2 \tab 1 \tab 4 \tab 3 \tab - \tab - \tab 2 \tab 5 \tab 5 \tab 23 \tab 2 \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 7 \tab - \tab 3 \tab 2 \tab - \tab 19 \tab 5 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 18 \tab 36 \tab 4 \tab - \tab - \tab 2 \tab - \tab 4 \tab 5 \tab 1 \tab 3 \tab 2 \tab 7 \tab 5 \tab 1 \tab 2 \tab 41 \tab 1 \tab 3 \tab 9 \tab 4 \tab - \tab 1 \tab 3 \tab 5 \tab - \tab - \tab 6 \tab 2 \tab 2 \tab 14 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 4 \tab 5 \tab 5 \tab 2 \tab 10 \tab 1 \tab 1 \tab 1 \tab 13 \tab 5 \tab 1 \tab 3 \tab 2 \tab - \tab 2 \tab 1 \tab 3 \tab 5 \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab 5 \tab - \tab - \tab - \tab 6 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 3 \tab 6 \tab 14 \tab 6 \tab 4 \tab 6 \tab 6 \tab 5 \tab 2 \tab 9 \tab 5 \tab 14 \tab 5 \tab 8 \tab 4 \tab - \tab 13 \tab 5 \tab 11 \tab 10 \tab 2 \tab 4 \tab 22 \tab 14 \tab 32 \tab 10 \tab 66 \tab 712 \tab 3 \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 4 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1 \tab 1 \tab 2 \tab 4 \tab 4 \tab 2 \tab - \tab - \tab 1 \tab - \tab 3 \tab 12 \tab 17 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 8 \tab - \tab 1 \tab 7 \tab 1 \tab 4 \tab - \tab 1 \tab 39 \tab 4 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab 3 \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 4 \tab 3 \tab 13 \tab 43 \tab 7 \tab - \tab 1 \tab 2 \tab - \tab 5 \tab 2 \tab 1 \tab 2 \tab 4 \tab 6 \tab 3 \tab - \tab 7 \tab 27 \tab - \tab 3 \tab 9 \tab 6 \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 4 \tab - \tab 1 \tab 10 \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 5 \tab 2 \tab 4 \tab 2 \tab 7 \tab 3 \tab 11 \tab 1 \tab 1 \tab - \tab 6 \tab 7 \tab 3 \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab 7 \tab 3 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 7 \tab 4 \tab 4 \tab 7 \tab 10 \tab 19 \tab 7 \tab - \tab 2 \tab 5 \tab 2 \tab 7 \tab 11 \tab 2 \tab 16 \tab 18 \tab 19 \tab 6 \tab 1 \tab 10 \tab 3 \tab 12 \tab 8 \tab 5 \tab 9 \tab 19 \tab 7 \tab 40 \tab 21 \tab 76 \tab 971 \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 6 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 6 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab 4 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 2 \tab 1 \tab 7 \tab 6 \tab 5 \tab 2 \tab 3 \tab - \tab - \tab 5 \tab 3 \tab 8 \tab 14 \tab - \tab 3 \tab - \tab - \tab - \tab 6 \tab - \tab - \tab 5 \tab - \tab 1 \tab 3 \tab 1 \tab 19 \tab 4 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 10 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 1 \tab 9 \tab 32 \tab 6 \tab - \tab 1 \tab 2 \tab 1 \tab 5 \tab 1 \tab 5 \tab 3 \tab 5 \tab 6 \tab 2 \tab 2 \tab 4 \tab 37 \tab - \tab 3 \tab 7 \tab 8 \tab 3 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 8 \tab 8 \tab - \tab 1 \tab 1 \tab 2 \tab 7 \tab - \tab 5 \tab 2 \tab 11 \tab 1 \tab 9 \tab 2 \tab - \tab 2 \tab 11 \tab 6 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab 5 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 11 \tab 1 \tab - \tab 1 \tab 7 \tab 1 \tab 3 \tab 7 \tab 4 \tab 5 \tab 8 \tab 7 \tab 12 \tab 8 \tab - \tab 3 \tab 7 \tab 5 \tab 1 \tab 13 \tab 5 \tab 16 \tab 5 \tab 19 \tab 2 \tab 2 \tab 5 \tab 4 \tab 13 \tab 6 \tab 4 \tab 4 \tab 15 \tab 10 \tab 26 \tab 26 \tab 41 \tab 749 \tab 1 \tab 3 \tab 2 \tab - \tab 3 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1481 \cr
   1989 \tab 2 \tab 1 \tab 9 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 8 \tab 6 \tab 3 \tab 15 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 6 \tab - \tab - \tab 5 \tab 2 \tab 3 \tab 7 \tab 3 \tab 11 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 9 \tab 26 \tab 2 \tab - \tab 1 \tab 4 \tab - \tab 4 \tab 12 \tab 6 \tab 3 \tab 2 \tab 9 \tab 2 \tab - \tab 4 \tab 48 \tab 2 \tab 4 \tab 13 \tab 7 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab 1 \tab 3 \tab 2 \tab 2 \tab 14 \tab 4 \tab 3 \tab 3 \tab 2 \tab - \tab 5 \tab 1 \tab 6 \tab 4 \tab 2 \tab 3 \tab 9 \tab 1 \tab 2 \tab 1 \tab 9 \tab 7 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab 4 \tab - \tab 6 \tab - \tab - \tab - \tab 9 \tab 2 \tab 1 \tab 6 \tab 2 \tab 5 \tab 8 \tab 10 \tab 15 \tab 8 \tab 9 \tab 2 \tab 1 \tab 3 \tab 2 \tab 12 \tab 6 \tab 21 \tab 11 \tab 19 \tab 4 \tab 1 \tab 10 \tab 4 \tab 19 \tab 4 \tab 5 \tab 9 \tab 15 \tab 7 \tab 30 \tab 18 \tab 52 \tab 753 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1537 \cr
   1990 \tab 1 \tab 5 \tab 7 \tab 2 \tab 6 \tab 1 \tab 3 \tab - \tab - \tab 3 \tab 4 \tab 1 \tab 14 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab - \tab 10 \tab 2 \tab 1 \tab 5 \tab 2 \tab 14 \tab 4 \tab - \tab 5 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 5 \tab - \tab - \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab 9 \tab 23 \tab 6 \tab - \tab 3 \tab - \tab - \tab 3 \tab 4 \tab 2 \tab 3 \tab 2 \tab 7 \tab 3 \tab 1 \tab 7 \tab 31 \tab 1 \tab 2 \tab 6 \tab 9 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 9 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 7 \tab 2 \tab 5 \tab 2 \tab 11 \tab 1 \tab 7 \tab - \tab - \tab - \tab 9 \tab 7 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 4 \tab 1 \tab 3 \tab 3 \tab 7 \tab - \tab 7 \tab 6 \tab 12 \tab 6 \tab 2 \tab 2 \tab 5 \tab 3 \tab 4 \tab 11 \tab 3 \tab 23 \tab 9 \tab 8 \tab 2 \tab 1 \tab 7 \tab 3 \tab 15 \tab 4 \tab 11 \tab 7 \tab 11 \tab 7 \tab 27 \tab 19 \tab 40 \tab 699 \tab - \tab 2 \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 16 \tab 42 \tab 71 \tab 49 \tab 59 \tab 20 \tab 52 \tab 5 \tab 12 \tab 55 \tab 48 \tab 162 \tab 306 \tab 55 \tab 17 \tab 15 \tab 3 \tab 1 \tab 93 \tab 6 \tab 7 \tab 99 \tab 18 \tab 49 \tab 52 \tab 21 \tab 331 \tab 64 \tab 6 \tab 41 \tab 25 \tab 27 \tab 4 \tab 6 \tab 18 \tab 18 \tab 43 \tab 13 \tab 17 \tab 76 \tab 8 \tab 25 \tab 26 \tab 8 \tab 8 \tab 7 \tab 14 \tab 22 \tab 11 \tab 41 \tab 32 \tab 287 \tab 561 \tab 105 \tab 6 \tab 15 \tab 34 \tab 8 \tab 55 \tab 77 \tab 69 \tab 123 \tab 50 \tab 140 \tab 69 \tab 13 \tab 76 \tab 669 \tab 16 \tab 54 \tab 168 \tab 129 \tab 15 \tab 29 \tab 33 \tab 28 \tab 29 \tab 5 \tab 34 \tab 9 \tab 19 \tab 211 \tab 93 \tab 17 \tab 22 \tab 16 \tab 44 \tab 62 \tab 36 \tab 56 \tab 43 \tab 154 \tab 39 \tab 176 \tab 18 \tab 9 \tab 22 \tab 188 \tab 125 \tab 28 \tab 37 \tab 10 \tab 76 \tab 23 \tab 10 \tab 42 \tab 51 \tab 25 \tab 42 \tab 49 \tab 57 \tab 24 \tab 87 \tab 7 \tab 1 \tab 2 \tab 103 \tab 11 \tab 40 \tab 101 \tab 80 \tab 41 \tab 105 \tab 165 \tab 221 \tab 123 \tab 45 \tab 74 \tab 104 \tab 88 \tab 50 \tab 163 \tab 97 \tab 298 \tab 216 \tab 361 \tab 59 \tab 17 \tab 176 \tab 103 \tab 367 \tab 120 \tab 141 \tab 179 \tab 321 \tab 150 \tab 680 \tab 418 \tab 1251 \tab 57999 \tab 16 \tab 20 \tab 14 \tab 4 \tab 49 \tab 8 \tab 28 \tab 3 \tab 21 \tab 12 \tab 15 \tab 5 \tab 10 \tab 6 \tab 7 \tab 2 \tab 23 \tab 8 \tab 17 \tab 10 \tab 17 \tab 12 \tab 14 \tab 7 \tab 32 \tab 14 \tab 13 \tab 4 \tab 12 \tab 6 \tab 29 \tab 22 \tab 14 \tab 38 \tab 16 \tab 14 \tab 7 \tab 15 \tab 5 \tab 21 \tab 8 \tab 10 \tab 16 \tab 4 \tab 4 \tab 1 \tab 4 \tab 3 \tab 9 \tab 5 \tab 13 \tab 3 \tab 13 \tab 2 \tab 5 \tab 6 \tab 14 \tab 2 \tab 3 \tab 12 \tab 12 \tab 33 \tab 4 \tab 13 \tab 7 \tab 11 \tab 15 \tab 4 \tab 5 \tab 10 \tab 7 \tab 2 \tab 3 \tab 1 \tab 3 \tab 3 \tab 1 \tab 5 \tab 3 \tab 1 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
