% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_574.R
\name{speth1}
\alias{speth1}
\title{speth1}
\source{
General Social Survey https://gss.norc.org
}
\description{
1st mentioned country of spouse's origin
}
\details{
Question 1598a. IF MORE THAN ONE COUNTRY NAMED: Which one of these countries does your (wife/husband) feel closer to?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab uncodeable \tab africa \tab american indian \tab american only \tab arabic \tab austria \tab china \tab czechoslovakia \tab denmark \tab don't know \tab england & wales \tab finland \tab france \tab french canada \tab germany \tab greece \tab hungary \tab india \tab ireland \tab italy \tab japan \tab lithuania \tab mexico \tab netherlands \tab no answer \tab non-span windies \tab norway \tab other \tab other canada \tab other spanish \tab philippines \tab poland \tab portugal \tab puerto rico \tab rumania \tab russia \tab scotland \tab spain \tab sweden \tab switzerland \tab yugoslavia \tab belgium \tab west indies \tab other asian \tab other european \tab not available in this year \tab iap \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 663 \tab 24 \tab 32 \tab 9 \tab 1 \tab 9 \tab 2 \tab 10 \tab 10 \tab 127 \tab 114 \tab 6 \tab 26 \tab 10 \tab 148 \tab 5 \tab 2 \tab 1 \tab 97 \tab 38 \tab 4 \tab 3 \tab 26 \tab 17 \tab 4 \tab 2 \tab 17 \tab 6 \tab 6 \tab 3 \tab 2 \tab 31 \tab 3 \tab 5 \tab 3 \tab 13 \tab 23 \tab 4 \tab 21 \tab 5 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 644 \tab 33 \tab 21 \tab 13 \tab 3 \tab 6 \tab - \tab 8 \tab 8 \tab 120 \tab 108 \tab 4 \tab 14 \tab 7 \tab 145 \tab 2 \tab 7 \tab 5 \tab 91 \tab 40 \tab 4 \tab 3 \tab 24 \tab 14 \tab 1 \tab 4 \tab 15 \tab 5 \tab 4 \tab 3 \tab 1 \tab 22 \tab 4 \tab 8 \tab 1 \tab 14 \tab 31 \tab 6 \tab 16 \tab 3 \tab 6 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 903 \tab 84 \tab 35 \tab 14 \tab 2 \tab 8 \tab 5 \tab 14 \tab 12 \tab 161 \tab 92 \tab 1 \tab 26 \tab 12 \tab 127 \tab 1 \tab 4 \tab 2 \tab 68 \tab 43 \tab 2 \tab 3 \tab 20 \tab 20 \tab 10 \tab 3 \tab 14 \tab 11 \tab 2 \tab 1 \tab 1 \tab 37 \tab 3 \tab 1 \tab 2 \tab 15 \tab 32 \tab 6 \tab 14 \tab 3 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 693 \tab 31 \tab 21 \tab 13 \tab - \tab 3 \tab 7 \tab 5 \tab 6 \tab 136 \tab 96 \tab 5 \tab 15 \tab 7 \tab 151 \tab 7 \tab 5 \tab 3 \tab 70 \tab 35 \tab 2 \tab 2 \tab 21 \tab 22 \tab 11 \tab 1 \tab 12 \tab 7 \tab 1 \tab 2 \tab - \tab 19 \tab - \tab 5 \tab 1 \tab 19 \tab 23 \tab 9 \tab 8 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 690 \tab 32 \tab 24 \tab 6 \tab 1 \tab 3 \tab 7 \tab 9 \tab 15 \tab 125 \tab 94 \tab 1 \tab 18 \tab 10 \tab 177 \tab 5 \tab 5 \tab 2 \tab 65 \tab 45 \tab 4 \tab 3 \tab 24 \tab 14 \tab 15 \tab 1 \tab 15 \tab 11 \tab 4 \tab 3 \tab 3 \tab 30 \tab 5 \tab 5 \tab 2 \tab 12 \tab 28 \tab 6 \tab 15 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 645 \tab 33 \tab 23 \tab 6 \tab 1 \tab 7 \tab 1 \tab 10 \tab 6 \tab 104 \tab 96 \tab 3 \tab 17 \tab 8 \tab 117 \tab 2 \tab 2 \tab 1 \tab 69 \tab 36 \tab 3 \tab - \tab 23 \tab 16 \tab 8 \tab 6 \tab 19 \tab 6 \tab 3 \tab 3 \tab - \tab 23 \tab 5 \tab 7 \tab 2 \tab 10 \tab 24 \tab 9 \tab 8 \tab 7 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 713 \tab 45 \tab 24 \tab 6 \tab - \tab 3 \tab 3 \tab 8 \tab 13 \tab 105 \tab 110 \tab 2 \tab 19 \tab 13 \tab 128 \tab 2 \tab 5 \tab 3 \tab 85 \tab 45 \tab 4 \tab 5 \tab 25 \tab 13 \tab 7 \tab 6 \tab 8 \tab 10 \tab 11 \tab 3 \tab 3 \tab 23 \tab 2 \tab 4 \tab - \tab 12 \tab 26 \tab 6 \tab 12 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 751 \tab 37 \tab 25 \tab 8 \tab 1 \tab 1 \tab 5 \tab 8 \tab 6 \tab 132 \tab 110 \tab 3 \tab 26 \tab 8 \tab 174 \tab 10 \tab 4 \tab 3 \tab 64 \tab 32 \tab 2 \tab 3 \tab 17 \tab 14 \tab 4 \tab 3 \tab 10 \tab 10 \tab 5 \tab 10 \tab 6 \tab 25 \tab 2 \tab 1 \tab 2 \tab 21 \tab 30 \tab 14 \tab 15 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2591 \tab 28 \tab 5 \tab 5 \tab 1 \tab 2 \tab - \tab 5 \tab 1 \tab 57 \tab 49 \tab 2 \tab 8 \tab 8 \tab 77 \tab 2 \tab 2 \tab 3 \tab 33 \tab 14 \tab 3 \tab 2 \tab 18 \tab 5 \tab 3 \tab - \tab 9 \tab 2 \tab 1 \tab 4 \tab 2 \tab 8 \tab 1 \tab 3 \tab 1 \tab 6 \tab 15 \tab - \tab 7 \tab 2 \tab 2 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 57779 \tab 347 \tab 210 \tab 80 \tab 10 \tab 42 \tab 30 \tab 77 \tab 77 \tab 1067 \tab 869 \tab 27 \tab 169 \tab 83 \tab 1244 \tab 36 \tab 36 \tab 23 \tab 642 \tab 328 \tab 28 \tab 24 \tab 198 \tab 135 \tab 63 \tab 26 \tab 119 \tab 68 \tab 37 \tab 32 \tab 18 \tab 218 \tab 25 \tab 39 \tab 14 \tab 122 \tab 232 \tab 60 \tab 116 \tab 30 \tab 21 \tab 7 \tab 1 \tab 4 \tab 1 \tab 4032 \tab 3544 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} africa
\item \code{2} austria
\item \code{3} french canada
\item \code{4} other canada
\item \code{5} china
\item \code{6} czechoslovakia
\item \code{7} denmark
\item \code{8} england & wales
\item \code{9} finland
\item \code{10} france
\item \code{11} germany
\item \code{12} greece
\item \code{13} hungary
\item \code{14} ireland
\item \code{15} italy
\item \code{16} japan
\item \code{17} mexico
\item \code{18} netherlands
\item \code{19} norway
\item \code{20} philippines
\item \code{21} poland
\item \code{22} puerto rico
\item \code{23} russia
\item \code{24} scotland
\item \code{25} spain
\item \code{26} sweden
\item \code{27} switzerland
\item \code{28} west indies
\item \code{29} other
\item \code{30} american indian
\item \code{31} india
\item \code{32} portugal
\item \code{33} lithuania
\item \code{34} yugoslavia
\item \code{35} rumania
\item \code{36} belgium
\item \code{37} arabic
\item \code{38} other spanish
\item \code{39} non-span windies
\item \code{40} other asian
\item \code{41} other european
\item \code{97} american only
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
