% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_574.R
\name{spethnic}
\alias{spethnic}
\title{spethnic}
\source{
General Social Survey https://gss.norc.org
}
\description{
Country of spouse's family origin
}
\details{
Question 1598. From what countries or part of the world did your (wife's/husband's) ancestors come?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab uncodeable \tab africa \tab american indian \tab american only \tab arabic \tab austria \tab china \tab czechoslovakia \tab denmark \tab england & wales \tab finland \tab france \tab french canada \tab germany \tab greece \tab india \tab ireland \tab italy \tab japan \tab lithuania \tab mexico \tab netherlands \tab no answer \tab non-span windies \tab norway \tab other \tab other canada \tab other spanish \tab philippines \tab poland \tab portugal \tab puerto rico \tab rumania \tab russia \tab scotland \tab spain \tab sweden \tab switzerland \tab yugoslavia \tab belgium \tab hungary \tab west indies \tab other asian \tab other european \tab not available in this year \tab iap \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 853 \tab 25 \tab 36 \tab 11 \tab 1 \tab 10 \tab 2 \tab 10 \tab 7 \tab 104 \tab 5 \tab 21 \tab 12 \tab 120 \tab 4 \tab 1 \tab 87 \tab 35 \tab 4 \tab 2 \tab 27 \tab 13 \tab 5 \tab 2 \tab 18 \tab 5 \tab 5 \tab 4 \tab 2 \tab 27 \tab 3 \tab 5 \tab 3 \tab 13 \tab 26 \tab 2 \tab 16 \tab 6 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 818 \tab 34 \tab 22 \tab 15 \tab 1 \tab 7 \tab - \tab 7 \tab 9 \tab 100 \tab 3 \tab 11 \tab 9 \tab 121 \tab 2 \tab 5 \tab 86 \tab 41 \tab 3 \tab 3 \tab 23 \tab 9 \tab 4 \tab 3 \tab 15 \tab 6 \tab 6 \tab 3 \tab 1 \tab 22 \tab 3 \tab 8 \tab 1 \tab 10 \tab 21 \tab 5 \tab 12 \tab 6 \tab 5 \tab 2 \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1134 \tab 84 \tab 34 \tab 14 \tab 2 \tab 7 \tab 4 \tab 14 \tab 10 \tab 78 \tab - \tab 17 \tab 13 \tab 111 \tab 1 \tab 2 \tab 63 \tab 47 \tab 2 \tab 2 \tab 23 \tab 13 \tab 10 \tab 3 \tab 15 \tab 11 \tab 2 \tab 1 \tab 1 \tab 30 \tab 3 \tab 1 \tab 2 \tab 12 \tab 24 \tab 3 \tab 15 \tab 3 \tab 2 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 880 \tab 29 \tab 28 \tab 13 \tab - \tab 3 \tab 7 \tab 7 \tab 4 \tab 85 \tab 5 \tab 12 \tab 10 \tab 116 \tab 7 \tab 3 \tab 67 \tab 35 \tab 2 \tab - \tab 22 \tab 25 \tab 10 \tab 1 \tab 10 \tab 8 \tab 3 \tab 2 \tab - \tab 17 \tab - \tab 5 \tab 1 \tab 15 \tab 18 \tab 9 \tab 10 \tab 5 \tab 2 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 888 \tab 32 \tab 26 \tab 6 \tab 1 \tab 2 \tab 7 \tab 8 \tab 11 \tab 87 \tab - \tab 13 \tab 12 \tab 139 \tab 4 \tab 2 \tab 75 \tab 43 \tab 4 \tab 3 \tab 25 \tab 11 \tab 15 \tab 1 \tab 15 \tab 12 \tab 4 \tab 3 \tab 3 \tab 28 \tab 3 \tab 5 \tab 1 \tab 6 \tab 24 \tab 5 \tab 9 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 802 \tab 34 \tab 26 \tab 6 \tab 1 \tab 6 \tab 1 \tab 12 \tab 7 \tab 80 \tab 3 \tab 13 \tab 7 \tab 99 \tab 2 \tab 1 \tab 64 \tab 32 \tab 3 \tab - \tab 23 \tab 14 \tab 13 \tab 6 \tab 17 \tab 6 \tab 4 \tab 4 \tab - \tab 21 \tab 3 \tab 7 \tab 1 \tab 8 \tab 21 \tab 7 \tab 7 \tab 5 \tab 4 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 879 \tab 46 \tab 24 \tab 6 \tab - \tab 2 \tab 3 \tab 7 \tab 12 \tab 88 \tab 1 \tab 17 \tab 15 \tab 109 \tab 2 \tab 3 \tab 78 \tab 46 \tab 4 \tab 3 \tab 25 \tab 10 \tab 12 \tab 6 \tab 8 \tab 13 \tab 9 \tab 2 \tab 3 \tab 22 \tab 1 \tab 5 \tab - \tab 10 \tab 15 \tab 5 \tab 14 \tab 4 \tab 2 \tab 1 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 939 \tab 37 \tab 28 \tab 9 \tab 1 \tab 1 \tab 3 \tab 6 \tab 6 \tab 103 \tab 2 \tab 16 \tab 6 \tab 143 \tab 8 \tab 3 \tab 64 \tab 33 \tab 2 \tab 2 \tab 13 \tab 13 \tab 16 \tab 3 \tab 9 \tab 14 \tab 4 \tab 9 \tab 8 \tab 23 \tab 2 \tab 1 \tab 2 \tab 18 \tab 21 \tab 15 \tab 16 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2659 \tab 26 \tab 7 \tab 5 \tab 1 \tab 2 \tab - \tab 5 \tab 1 \tab 38 \tab 2 \tab 6 \tab 10 \tab 71 \tab 2 \tab 3 \tab 33 \tab 15 \tab 3 \tab 2 \tab 19 \tab 6 \tab 13 \tab - \tab 7 \tab 3 \tab 2 \tab 3 \tab 2 \tab 7 \tab 1 \tab 2 \tab 1 \tab 4 \tab 13 \tab - \tab 8 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 4 \tab 1 \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 59338 \tab 347 \tab 231 \tab 85 \tab 8 \tab 40 \tab 27 \tab 76 \tab 67 \tab 763 \tab 21 \tab 126 \tab 94 \tab 1029 \tab 32 \tab 23 \tab 617 \tab 327 \tab 27 \tab 17 \tab 200 \tab 114 \tab 98 \tab 25 \tab 114 \tab 78 \tab 39 \tab 31 \tab 20 \tab 197 \tab 19 \tab 39 \tab 12 \tab 96 \tab 183 \tab 51 \tab 107 \tab 35 \tab 20 \tab 5 \tab 30 \tab 1 \tab 4 \tab 1 \tab 4032 \tab 3544 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} africa
\item \code{2} austria
\item \code{3} french canada
\item \code{4} other canada
\item \code{5} china
\item \code{6} czechoslovakia
\item \code{7} denmark
\item \code{8} england & wales
\item \code{9} finland
\item \code{10} france
\item \code{11} germany
\item \code{12} greece
\item \code{13} hungary
\item \code{14} ireland
\item \code{15} italy
\item \code{16} japan
\item \code{17} mexico
\item \code{18} netherlands
\item \code{19} norway
\item \code{20} philippines
\item \code{21} poland
\item \code{22} puerto rico
\item \code{23} russia
\item \code{24} scotland
\item \code{25} spain
\item \code{26} sweden
\item \code{27} switzerland
\item \code{28} west indies
\item \code{29} other
\item \code{30} american indian
\item \code{31} india
\item \code{32} portugal
\item \code{33} lithuania
\item \code{34} yugoslavia
\item \code{35} rumania
\item \code{36} belgium
\item \code{37} arabic
\item \code{38} other spanish
\item \code{39} non-span windies
\item \code{40} other asian
\item \code{41} other european
\item \code{97} american only
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
