% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_04.R
\name{sphrs1}
\alias{sphrs1}
\title{Number of hrs spouse worked last week}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphrs1
}
\details{
Question 5a. If working, full or part time: How many hours did (he/she) work last week, at all jobs?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 10 \tab 11 \tab 12 \tab 15 \tab 16 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 24 \tab 25 \tab 28 \tab 3 \tab 30 \tab 32 \tab 33 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 42 \tab 43 \tab 44 \tab 45 \tab 48 \tab 5 \tab 50 \tab 51 \tab 52 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 62 \tab 63 \tab 65 \tab 66 \tab 70 \tab 72 \tab 74 \tab 75 \tab 79 \tab 8 \tab 80 \tab 84 \tab 89+ hours \tab don't know \tab no answer \tab 13 \tab 23 \tab 27 \tab 29 \tab 34 \tab 46 \tab 47 \tab 6 \tab 68 \tab 69 \tab 73 \tab 9 \tab 14 \tab 41 \tab 49 \tab 64 \tab 76 \tab 78 \tab 2 \tab 26 \tab 53 \tab 77 \tab 82 \tab 85 \tab 31 \tab 61 \tab 88 \tab 7 \tab 17 \tab 71 \tab 0 \tab 67 \tab 81 \tab 1 \tab 86 \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 853 \tab 3 \tab 1 \tab 4 \tab 5 \tab 7 \tab 3 \tab 2 \tab 10 \tab 1 \tab 1 \tab 7 \tab 7 \tab 2 \tab 3 \tab 20 \tab 11 \tab 1 \tab 14 \tab 8 \tab 6 \tab 9 \tab 2 \tab 1 \tab 282 \tab 3 \tab 5 \tab 10 \tab 29 \tab 31 \tab 1 \tab 52 \tab 3 \tab 5 \tab 3 \tab 11 \tab 7 \tab 1 \tab 1 \tab 1 \tab 28 \tab 2 \tab 2 \tab 9 \tab 1 \tab 8 \tab 4 \tab 2 \tab 2 \tab 1 \tab 2 \tab 3 \tab 4 \tab 4 \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 849 \tab 3 \tab - \tab 2 \tab 6 \tab 3 \tab 1 \tab - \tab 13 \tab 2 \tab 1 \tab 6 \tab 5 \tab 2 \tab - \tab 15 \tab 10 \tab - \tab 18 \tab 3 \tab 10 \tab 6 \tab - \tab 1 \tab 294 \tab 5 \tab 2 \tab 19 \tab 28 \tab 32 \tab 1 \tab 38 \tab 2 \tab 1 \tab 7 \tab 9 \tab 6 \tab - \tab - \tab - \tab 27 \tab 1 \tab - \tab 3 \tab 2 \tab 8 \tab 5 \tab - \tab 1 \tab - \tab 7 \tab 1 \tab - \tab 5 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 896 \tab 9 \tab - \tab 3 \tab 7 \tab 9 \tab 3 \tab 1 \tab 16 \tab 2 \tab 1 \tab 10 \tab 6 \tab 3 \tab 1 \tab 8 \tab 6 \tab 1 \tab 16 \tab 2 \tab 8 \tab 5 \tab 1 \tab - \tab 307 \tab 6 \tab - \tab 7 \tab 25 \tab 19 \tab - \tab 30 \tab - \tab 3 \tab 1 \tab 5 \tab 3 \tab - \tab 1 \tab - \tab 26 \tab - \tab - \tab 5 \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 5 \tab - \tab 6 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 942 \tab 1 \tab 1 \tab 2 \tab 4 \tab 6 \tab 2 \tab - \tab 17 \tab 1 \tab 2 \tab 11 \tab 10 \tab 1 \tab - \tab 14 \tab 4 \tab - \tab 23 \tab 5 \tab 10 \tab 8 \tab - \tab - \tab 245 \tab 5 \tab 3 \tab 7 \tab 22 \tab 24 \tab 1 \tab 32 \tab 3 \tab 2 \tab 2 \tab 4 \tab 4 \tab - \tab 3 \tab 2 \tab 32 \tab 3 \tab - \tab 1 \tab - \tab 8 \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 5 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 953 \tab 4 \tab - \tab 4 \tab 3 \tab 5 \tab 1 \tab - \tab 14 \tab 2 \tab 2 \tab 6 \tab 9 \tab 1 \tab - \tab 10 \tab 9 \tab 1 \tab 25 \tab 11 \tab 8 \tab 13 \tab 1 \tab 1 \tab 265 \tab 4 \tab 2 \tab 5 \tab 20 \tab 13 \tab - \tab 30 \tab 1 \tab 3 \tab 5 \tab 8 \tab 6 \tab - \tab 1 \tab - \tab 28 \tab - \tab - \tab 6 \tab - \tab 6 \tab 2 \tab 1 \tab 3 \tab - \tab 10 \tab 4 \tab 2 \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 943 \tab 11 \tab 1 \tab 3 \tab 3 \tab 10 \tab 1 \tab - \tab 11 \tab - \tab 2 \tab 9 \tab 8 \tab 4 \tab 2 \tab 13 \tab 11 \tab 1 \tab 10 \tab 3 \tab 2 \tab 7 \tab 1 \tab 1 \tab 254 \tab 5 \tab 2 \tab 5 \tab 22 \tab 36 \tab 2 \tab 39 \tab 2 \tab - \tab 4 \tab 7 \tab 4 \tab - \tab - \tab - \tab 27 \tab 1 \tab 1 \tab 4 \tab - \tab 5 \tab 8 \tab - \tab - \tab - \tab 6 \tab - \tab 1 \tab 16 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 923 \tab 5 \tab - \tab 5 \tab 6 \tab 10 \tab 1 \tab 1 \tab 11 \tab 2 \tab 1 \tab 7 \tab 9 \tab 2 \tab 1 \tab 10 \tab 7 \tab 1 \tab 12 \tab 1 \tab 3 \tab 5 \tab - \tab - \tab 233 \tab 6 \tab - \tab 11 \tab 27 \tab 21 \tab 1 \tab 40 \tab - \tab 1 \tab 2 \tab 10 \tab 4 \tab 1 \tab 2 \tab - \tab 29 \tab 2 \tab - \tab 5 \tab 1 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab 3 \tab 8 \tab 2 \tab 6 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1246 \tab 4 \tab - \tab 7 \tab 3 \tab 7 \tab 2 \tab 1 \tab 16 \tab 2 \tab 1 \tab 6 \tab 7 \tab 4 \tab 1 \tab 14 \tab 15 \tab 2 \tab 13 \tab 5 \tab 14 \tab 6 \tab - \tab - \tab 304 \tab 5 \tab 2 \tab 5 \tab 15 \tab 19 \tab - \tab 38 \tab 1 \tab 2 \tab 4 \tab 7 \tab 4 \tab 1 \tab 1 \tab - \tab 19 \tab - \tab 1 \tab 4 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 6 \tab 5 \tab 2 \tab 3 \tab 5 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1008 \tab 7 \tab 1 \tab 5 \tab 4 \tab 6 \tab 1 \tab - \tab 13 \tab - \tab 2 \tab 5 \tab 4 \tab 2 \tab 2 \tab 24 \tab 11 \tab - \tab 18 \tab 7 \tab 6 \tab 7 \tab 1 \tab 3 \tab 276 \tab 6 \tab 2 \tab 4 \tab 20 \tab 13 \tab 2 \tab 38 \tab - \tab 2 \tab 1 \tab 12 \tab 4 \tab - \tab 1 \tab - \tab 26 \tab 2 \tab - \tab 8 \tab - \tab 6 \tab - \tab - \tab - \tab - \tab 4 \tab 6 \tab 1 \tab 2 \tab 2 \tab 6 \tab 2 \tab - \tab 1 \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 946 \tab 3 \tab - \tab 2 \tab 3 \tab 5 \tab 1 \tab - \tab 16 \tab - \tab 1 \tab 6 \tab 4 \tab - \tab 1 \tab 11 \tab 9 \tab - \tab 11 \tab 4 \tab 2 \tab 7 \tab - \tab - \tab 267 \tab 5 \tab 1 \tab 4 \tab 16 \tab 19 \tab - \tab 36 \tab - \tab 3 \tab 1 \tab 7 \tab 4 \tab 1 \tab 2 \tab - \tab 18 \tab - \tab - \tab 7 \tab 1 \tab 9 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 3 \tab 4 \tab - \tab 7 \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 999 \tab 5 \tab - \tab 3 \tab 7 \tab 7 \tab - \tab - \tab 17 \tab - \tab 2 \tab 6 \tab 4 \tab 4 \tab 1 \tab 12 \tab 12 \tab - \tab 11 \tab 3 \tab 3 \tab 13 \tab 1 \tab 1 \tab 233 \tab 6 \tab 1 \tab 6 \tab 25 \tab 14 \tab 1 \tab 59 \tab 1 \tab 3 \tab 4 \tab 9 \tab 4 \tab 1 \tab 1 \tab - \tab 18 \tab 1 \tab - \tab - \tab - \tab 8 \tab 5 \tab - \tab - \tab 1 \tab 4 \tab 2 \tab 2 \tab 5 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 954 \tab 4 \tab - \tab 5 \tab 5 \tab 7 \tab 2 \tab - \tab 25 \tab - \tab 1 \tab 12 \tab 3 \tab 3 \tab - \tab 22 \tab 6 \tab 1 \tab 11 \tab 4 \tab 5 \tab 7 \tab - \tab 3 \tab 196 \tab 5 \tab 2 \tab 5 \tab 19 \tab 15 \tab 2 \tab 34 \tab 3 \tab 3 \tab 2 \tab 11 \tab 1 \tab - \tab - \tab - \tab 32 \tab - \tab - \tab 2 \tab 2 \tab 8 \tab 1 \tab 1 \tab 6 \tab - \tab 3 \tab 7 \tab - \tab 2 \tab 5 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 3 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1202 \tab 3 \tab 1 \tab 4 \tab 8 \tab 7 \tab - \tab 1 \tab 11 \tab 1 \tab 2 \tab 8 \tab 5 \tab 2 \tab - \tab 12 \tab 11 \tab - \tab 19 \tab 5 \tab 4 \tab 7 \tab 2 \tab 2 \tab 287 \tab 5 \tab 5 \tab 3 \tab 30 \tab 16 \tab 1 \tab 38 \tab - \tab 4 \tab 2 \tab 13 \tab 2 \tab 1 \tab 2 \tab 1 \tab 30 \tab 1 \tab 1 \tab 8 \tab 1 \tab 8 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 5 \tab - \tab 16 \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1000 \tab 1 \tab - \tab 1 \tab 3 \tab 9 \tab 2 \tab - \tab 11 \tab 1 \tab - \tab 8 \tab 4 \tab - \tab - \tab 21 \tab 9 \tab - \tab 8 \tab 4 \tab 2 \tab 6 \tab - \tab - \tab 209 \tab 2 \tab 6 \tab 4 \tab 19 \tab 8 \tab - \tab 45 \tab - \tab 4 \tab 4 \tab 9 \tab 2 \tab - \tab 1 \tab - \tab 38 \tab 2 \tab - \tab 1 \tab - \tab 6 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 7 \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 990 \tab 3 \tab - \tab 1 \tab 4 \tab 6 \tab 1 \tab - \tab 14 \tab - \tab - \tab 6 \tab 8 \tab 1 \tab 1 \tab 19 \tab 12 \tab - \tab 13 \tab 5 \tab 3 \tab 4 \tab - \tab 1 \tab 235 \tab 3 \tab 1 \tab 7 \tab 31 \tab 14 \tab - \tab 43 \tab - \tab 5 \tab 1 \tab 16 \tab 4 \tab - \tab 1 \tab - \tab 29 \tab 2 \tab - \tab 4 \tab 2 \tab 7 \tab - \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab 14 \tab - \tab - \tab 3 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 913 \tab 2 \tab 1 \tab 2 \tab 5 \tab 3 \tab 2 \tab 1 \tab 16 \tab - \tab 1 \tab 7 \tab 6 \tab 1 \tab - \tab 13 \tab 6 \tab - \tab 10 \tab 1 \tab 2 \tab 7 \tab 1 \tab - \tab 211 \tab 5 \tab 1 \tab 5 \tab 21 \tab 13 \tab 1 \tab 30 \tab 1 \tab 1 \tab 1 \tab 9 \tab 2 \tab - \tab - \tab - \tab 31 \tab - \tab - \tab 2 \tab 1 \tab 7 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 6 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1013 \tab 2 \tab - \tab 3 \tab 5 \tab 8 \tab 2 \tab - \tab 13 \tab 1 \tab - \tab 8 \tab 2 \tab 1 \tab - \tab 13 \tab 8 \tab - \tab 8 \tab 3 \tab 3 \tab 5 \tab 1 \tab 1 \tab 221 \tab 4 \tab 2 \tab 4 \tab 28 \tab 12 \tab - \tab 40 \tab - \tab 2 \tab 2 \tab 4 \tab 1 \tab - \tab 3 \tab - \tab 45 \tab - \tab - \tab 6 \tab - \tab 6 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 4 \tab 1 \tab 5 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 5 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1036 \tab 6 \tab - \tab 3 \tab 4 \tab 6 \tab 1 \tab - \tab 20 \tab - \tab 2 \tab 8 \tab 11 \tab - \tab 2 \tab 10 \tab 6 \tab - \tab 9 \tab 2 \tab 2 \tab 10 \tab 1 \tab 1 \tab 223 \tab 6 \tab 3 \tab 4 \tab 38 \tab 11 \tab 2 \tab 48 \tab - \tab 3 \tab 1 \tab 16 \tab 1 \tab 1 \tab 1 \tab - \tab 44 \tab 1 \tab - \tab 9 \tab - \tab 8 \tab 5 \tab - \tab 1 \tab - \tab 6 \tab 5 \tab - \tab 7 \tab - \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1969 \tab 7 \tab - \tab 7 \tab 9 \tab 5 \tab 2 \tab - \tab 40 \tab 1 \tab 1 \tab 15 \tab 14 \tab 7 \tab 2 \tab 31 \tab 11 \tab 2 \tab 18 \tab 12 \tab 8 \tab 13 \tab 2 \tab 1 \tab 415 \tab 4 \tab 5 \tab 8 \tab 42 \tab 26 \tab 1 \tab 97 \tab - \tab 6 \tab 4 \tab 28 \tab 9 \tab - \tab 2 \tab - \tab 65 \tab - \tab 1 \tab 11 \tab - \tab 16 \tab 5 \tab - \tab 2 \tab - \tab 3 \tab 13 \tab 3 \tab 4 \tab 5 \tab 6 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 8 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1915 \tab 8 \tab - \tab 3 \tab 6 \tab 6 \tab 2 \tab 1 \tab 38 \tab 3 \tab 4 \tab 17 \tab 15 \tab 2 \tab 2 \tab 26 \tab 9 \tab 1 \tab 28 \tab 14 \tab 6 \tab 11 \tab - \tab 2 \tab 399 \tab 7 \tab 1 \tab 2 \tab 42 \tab 24 \tab 2 \tab 94 \tab - \tab 8 \tab 3 \tab 26 \tab 6 \tab - \tab 3 \tab - \tab 70 \tab - \tab 4 \tab 7 \tab 1 \tab 19 \tab 3 \tab 1 \tab 4 \tab - \tab 6 \tab 10 \tab 1 \tab 3 \tab 2 \tab 10 \tab - \tab 1 \tab - \tab - \tab 2 \tab 5 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1888 \tab 8 \tab - \tab 3 \tab 7 \tab 3 \tab 2 \tab 1 \tab 24 \tab 2 \tab 1 \tab 11 \tab 14 \tab - \tab 2 \tab 27 \tab 8 \tab - \tab 22 \tab 5 \tab 3 \tab 11 \tab 2 \tab 1 \tab 399 \tab 9 \tab 4 \tab 8 \tab 43 \tab 22 \tab 1 \tab 98 \tab - \tab 6 \tab - \tab 27 \tab 6 \tab 2 \tab 2 \tab - \tab 57 \tab 2 \tab 2 \tab 6 \tab - \tab 18 \tab 5 \tab 2 \tab - \tab - \tab 8 \tab 8 \tab - \tab 5 \tab - \tab 13 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 4 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1923 \tab 9 \tab - \tab 5 \tab 6 \tab 6 \tab 6 \tab - \tab 24 \tab 1 \tab 1 \tab 9 \tab 12 \tab 2 \tab - \tab 30 \tab 11 \tab - \tab 15 \tab 8 \tab 5 \tab 8 \tab - \tab - \tab 381 \tab 12 \tab 2 \tab 2 \tab 33 \tab 21 \tab 1 \tab 103 \tab 1 \tab 3 \tab 6 \tab 16 \tab 6 \tab 1 \tab 2 \tab - \tab 56 \tab - \tab - \tab 7 \tab 1 \tab 18 \tab 3 \tab - \tab - \tab - \tab 1 \tab 13 \tab 1 \tab 6 \tab 2 \tab 13 \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1917 \tab 9 \tab - \tab 3 \tab 10 \tab 13 \tab 3 \tab - \tab 28 \tab 1 \tab 4 \tab 13 \tab 16 \tab 3 \tab - \tab 22 \tab 13 \tab - \tab 19 \tab 13 \tab 3 \tab 5 \tab 1 \tab - \tab 333 \tab 8 \tab 1 \tab 5 \tab 46 \tab 19 \tab 1 \tab 96 \tab - \tab 5 \tab - \tab 18 \tab 7 \tab 3 \tab 2 \tab - \tab 42 \tab 1 \tab 1 \tab 8 \tab - \tab 13 \tab 3 \tab - \tab 4 \tab - \tab 4 \tab 7 \tab - \tab 1 \tab 7 \tab 8 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 6 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab 1828 \tab 4 \tab 1 \tab - \tab 5 \tab 4 \tab 2 \tab 1 \tab 23 \tab - \tab 2 \tab 11 \tab 14 \tab 4 \tab - \tab 25 \tab 14 \tab - \tab 17 \tab 11 \tab 3 \tab 7 \tab - \tab 2 \tab 405 \tab 11 \tab 3 \tab 4 \tab 51 \tab 12 \tab 2 \tab 98 \tab 1 \tab 6 \tab 2 \tab 21 \tab 4 \tab - \tab - \tab - \tab 73 \tab 3 \tab - \tab 14 \tab - \tab 16 \tab 5 \tab - \tab 3 \tab - \tab 7 \tab 19 \tab 1 \tab 6 \tab 15 \tab 19 \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2812 \cr
   2006 \tab 3076 \tab 11 \tab - \tab 8 \tab 9 \tab 12 \tab 3 \tab 1 \tab 46 \tab 1 \tab - \tab 15 \tab 17 \tab 5 \tab 3 \tab 50 \tab 29 \tab 1 \tab 20 \tab 22 \tab 5 \tab 11 \tab 5 \tab - \tab 601 \tab 12 \tab 6 \tab 11 \tab 80 \tab 25 \tab 4 \tab 130 \tab 1 \tab 7 \tab 7 \tab 34 \tab 6 \tab 2 \tab 3 \tab - \tab 95 \tab 1 \tab - \tab 12 \tab 3 \tab 21 \tab 4 \tab - \tab 3 \tab - \tab 1 \tab 21 \tab 1 \tab 6 \tab 12 \tab 14 \tab - \tab 1 \tab - \tab - \tab 4 \tab 6 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 1381 \tab 3 \tab - \tab 2 \tab 6 \tab 6 \tab 5 \tab 1 \tab 21 \tab - \tab 1 \tab 2 \tab 16 \tab 1 \tab - \tab 21 \tab 9 \tab 1 \tab 18 \tab 7 \tab 3 \tab 1 \tab - \tab 1 \tab 251 \tab 5 \tab 4 \tab 3 \tab 25 \tab 13 \tab - \tab 82 \tab - \tab 6 \tab - \tab 15 \tab 3 \tab - \tab 2 \tab - \tab 38 \tab 2 \tab - \tab 3 \tab 1 \tab 8 \tab 5 \tab - \tab 2 \tab - \tab - \tab 8 \tab 1 \tab 2 \tab 8 \tab 11 \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab 1507 \tab 5 \tab - \tab 3 \tab 4 \tab 3 \tab 2 \tab - \tab 26 \tab - \tab 2 \tab 7 \tab 6 \tab - \tab 2 \tab 21 \tab 9 \tab 1 \tab 11 \tab 7 \tab 3 \tab 3 \tab 1 \tab 2 \tab 213 \tab 9 \tab 1 \tab 3 \tab 25 \tab 10 \tab - \tab 45 \tab - \tab 5 \tab - \tab 13 \tab 1 \tab - \tab 1 \tab - \tab 34 \tab - \tab - \tab 4 \tab - \tab 9 \tab 2 \tab - \tab 2 \tab - \tab 5 \tab 6 \tab - \tab 3 \tab 3 \tab 11 \tab - \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 2044 \cr
   2012 \tab 1402 \tab 6 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 16 \tab - \tab 1 \tab 6 \tab 5 \tab - \tab - \tab 21 \tab 10 \tab 1 \tab 10 \tab 5 \tab 5 \tab 6 \tab 2 \tab 3 \tab 237 \tab 4 \tab 2 \tab 5 \tab 32 \tab 7 \tab 1 \tab 57 \tab - \tab 5 \tab 2 \tab 11 \tab - \tab 1 \tab - \tab - \tab 45 \tab - \tab - \tab 6 \tab - \tab 6 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 6 \tab 2 \tab 3 \tab 6 \tab 6 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab 1805 \tab 5 \tab - \tab 6 \tab 9 \tab 3 \tab 3 \tab - \tab 24 \tab 2 \tab 1 \tab 11 \tab 10 \tab 3 \tab - \tab 27 \tab 11 \tab - \tab 15 \tab 5 \tab 2 \tab - \tab - \tab - \tab 292 \tab 6 \tab 1 \tab 6 \tab 40 \tab 13 \tab 1 \tab 73 \tab 1 \tab 2 \tab 1 \tab 15 \tab 3 \tab - \tab - \tab - \tab 55 \tab - \tab - \tab 10 \tab - \tab 22 \tab 3 \tab - \tab 3 \tab - \tab 5 \tab 6 \tab 2 \tab 2 \tab 7 \tab 7 \tab - \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 2538 \cr
   2016 \tab 2103 \tab 3 \tab - \tab 2 \tab 10 \tab 2 \tab 3 \tab - \tab 19 \tab - \tab - \tab 6 \tab 3 \tab 3 \tab - \tab 22 \tab 16 \tab 1 \tab 14 \tab 11 \tab 5 \tab 3 \tab 5 \tab 2 \tab 302 \tab 4 \tab 2 \tab 3 \tab 33 \tab 18 \tab 3 \tab 101 \tab 1 \tab 3 \tab 1 \tab 21 \tab 1 \tab - \tab 2 \tab - \tab 58 \tab 1 \tab - \tab 6 \tab - \tab 10 \tab 1 \tab - \tab 3 \tab - \tab 5 \tab 8 \tab 1 \tab 6 \tab 7 \tab 14 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 2867 \cr
   2018 \tab 1699 \tab 6 \tab - \tab - \tab 7 \tab 8 \tab 1 \tab 1 \tab 19 \tab 1 \tab 3 \tab 12 \tab 6 \tab 5 \tab - \tab 18 \tab 6 \tab - \tab 17 \tab 4 \tab 3 \tab 4 \tab 1 \tab 2 \tab 263 \tab 2 \tab 2 \tab 3 \tab 41 \tab 16 \tab 1 \tab 66 \tab - \tab 1 \tab 1 \tab 10 \tab 1 \tab - \tab 1 \tab - \tab 54 \tab 1 \tab - \tab 7 \tab - \tab 6 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 12 \tab 2 \tab 1 \tab 5 \tab 7 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 2348 \cr
   2021 \tab 2914 \tab 7 \tab - \tab 3 \tab 11 \tab 7 \tab 1 \tab - \tab 35 \tab 2 \tab - \tab 20 \tab 15 \tab 4 \tab - \tab 50 \tab 16 \tab - \tab 23 \tab 8 \tab - \tab 6 \tab 2 \tab 2 \tab 511 \tab 9 \tab 6 \tab 6 \tab 63 \tab 19 \tab 2 \tab 114 \tab - \tab 7 \tab - \tab 14 \tab 7 \tab - \tab 2 \tab 1 \tab 52 \tab - \tab - \tab 10 \tab 3 \tab 8 \tab 3 \tab - \tab 2 \tab - \tab 6 \tab 8 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab 1 \tab - \tab 22 \tab 4032 \cr
   2022 \tab 2664 \tab 12 \tab - \tab 2 \tab 6 \tab 5 \tab 3 \tab - \tab 30 \tab 1 \tab 3 \tab 15 \tab 10 \tab 4 \tab - \tab 31 \tab 16 \tab - \tab 21 \tab 6 \tab 6 \tab 4 \tab - \tab 2 \tab 391 \tab 7 \tab 2 \tab 2 \tab 48 \tab 10 \tab 2 \tab 95 \tab - \tab 3 \tab 2 \tab 10 \tab 2 \tab 1 \tab 1 \tab - \tab 48 \tab - \tab - \tab 5 \tab 1 \tab 7 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 9 \tab - \tab 5 \tab 6 \tab 10 \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab 7 \tab 3544 \cr
   Total \tab 49270 \tab 179 \tab 8 \tab 109 \tab 191 \tab 207 \tab 65 \tab 13 \tab 677 \tab 30 \tab 46 \tab 306 \tab 285 \tab 76 \tab 26 \tab 683 \tab 351 \tab 16 \tab 517 \tab 214 \tab 153 \tab 225 \tab 33 \tab 36 \tab 9935 \tab 195 \tab 82 \tab 186 \tab 1079 \tab 585 \tab 37 \tab 2059 \tab 22 \tab 120 \tab 76 \tab 446 \tab 125 \tab 17 \tab 44 \tab 5 \tab 1369 \tab 29 \tab 13 \tab 200 \tab 23 \tab 316 \tab 95 \tab 8 \tab 58 \tab 2 \tab 124 \tab 212 \tab 37 \tab 134 \tab 115 \tab 233 \tab 12 \tab 14 \tab 36 \tab 14 \tab 36 \tab 106 \tab 50 \tab 50 \tab 22 \tab 4 \tab 4 \tab 13 \tab 20 \tab 29 \tab 39 \tab 18 \tab 9 \tab 5 \tab 14 \tab 28 \tab 19 \tab 7 \tab 3 \tab 12 \tab 11 \tab 7 \tab 6 \tab 14 \tab 12 \tab 4 \tab 17 \tab 5 \tab 1 \tab 22 \tab 1 \tab 29 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{89} 89+ hours
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
