% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_06.R
\name{spind10}
\alias{spind10}
\title{Spouse's industry code (naics 2007)}
\source{
General Social Survey https://gss.norc.org
}
\description{
spind10
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab accounting, tax preparation, bookkeeping, and payroll services \tab advertising and related services \tab aerospace product and parts manufacturing \tab agricultural chemical manufacturing \tab agricultural implement manufacturing \tab air transportation \tab aircraft and parts manufacturing \tab animal production \tab animal slaughtering and processing \tab apparel, fabrics, and notions, merchant wholesalers \tab architectural, engineering, and related services \tab auto parts, accessories, and tire stores \tab automobile dealers \tab automotive repair and maintenance \tab banking and related activities \tab barber shops \tab beauty salons \tab beverage manufacturing \tab bowling centers \tab building material and supplies dealers \tab bus service and urban transit \tab business support services \tab business, professional, political, and similar organizations \tab business, technical, and trade schools and training \tab cement, concrete, lime, and gypsum product manufacturing \tab child day care services \tab civic, social, advocacy organizations, and grantmaking and giving services \tab clothing stores \tab colleges and universities, including junior colleges \tab communications, and audio and video equipment manufacturing \tab computer and peripheral equipment manufacturing \tab construction \tab construction, and mining and oil and gas field machinery manufacturing \tab crop production \tab cut and sew apparel manufacturing \tab dairy product manufacturing \tab department stores and discount stores \tab drugs, sundries, and chemical and allied products, merchant wholesalers \tab drycleaning and laundry services \tab electric and gas, and other combinations \tab electric power generation, transmission, and distribution \tab electrical and electronic goods, merchant wholesalers \tab electrical lighting and electrical equipment manufacturing, and other electrical component manufacturing, n.e.c. \tab electronic component and product manufacturing, n.e.c. \tab elementary and secondary schools \tab employment services \tab executive offices and legislative bodies \tab fabric mills, except knitting mills \tab farm product raw materials, merchant wholesalers \tab fiber, yarn, and thread mills \tab footwear manufacturing \tab forestry, except logging \tab fruit and vegetable preserving and specialty food manufacturing \tab fuel dealers \tab furniture and home furnishings stores \tab furniture and related product manufacturing \tab gasoline stations \tab glass and glass product manufacturing \tab groceries and related products, merchant wholesalers \tab grocery stores \tab hardware, plumbing and heating equipment, and supplies, merchant wholesalers \tab health and personal care, except drug, stores \tab hospitals \tab household appliance manufacturing \tab household appliance stores \tab iap \tab independent artists, performing arts, spectator sports, and related industries \tab individual and family services \tab industrial and miscellaneous chemicals \tab insurance carriers and related activities \tab iron and steel mills and steel product manufacturing \tab jewelry, luggage, and leather goods stores \tab justice, public order, and safety activities \tab lawn and garden equipment and supplies stores \tab leather tanning and finishing and other allied products manufacturing \tab legal services \tab libraries and archives \tab machine shops; turned product; screw, nut, and bolt manufacturing \tab machinery manufacturing, n.e.c. \tab machinery, equipment, and supplies, merchant wholesalers \tab management of companies and enterprises \tab management, scientific, and technical consulting services \tab medical equipment and supplies manufacturing \tab metal ore mining \tab metals and minerals, except petroleum, merchant wholesalers \tab metalworking machinery manufacturing \tab miscellaneous fabricated metal products manufacturing \tab miscellaneous general merchandise stores \tab miscellaneous manufacturing, n.e.c. \tab miscellaneous nonmetallic mineral product manufacturing \tab miscellaneous paper and pulp products \tab miscellaneous retail stores \tab miscellaneous wood products \tab motor vehicles and motor vehicle equipment manufacturing \tab national security and international affairs \tab natural gas distribution \tab newspaper publishers \tab no answer \tab non-depository credit and related activities \tab nonferrous metal (except aluminum) production and processing \tab not specified food industries \tab not specified machinery manufacturing \tab not specified manufacturing industries \tab not specified metal industries \tab not specified wholesale trade \tab nursing care facilities \tab office supplies and stationery stores \tab offices of dentists \tab offices of optometrists \tab offices of physicians \tab ordnance \tab other administrative and other support services \tab other amusement, gambling, and recreation industries \tab other direct selling establishments \tab other general government and support \tab other motor vehicle dealers \tab other personal services \tab other professional, scientific, and technical services \tab other schools and instruction, and educational support services \tab paper and paper products, merchant wholesalers \tab paperboard containers and boxes \tab periodical, book, and directory publishers \tab personal and household goods repair and maintenance \tab petroleum and petroleum products, merchant wholesalers \tab petroleum refining \tab pharmaceutical and medicine manufacturing \tab pharmacies and drug stores \tab postal service \tab pottery, ceramics, and plumbing fixture manufacturing \tab printing and related support activities \tab private households \tab professional and commercial equipment and supplies, merchant wholesalers \tab public finance activities \tab pulp, paper, and paperboard mills \tab radio and television broadcasting and cable subscription programming \tab radio, tv, and computer stores \tab rail transportation \tab railroad rolling stock manufacturing \tab real estate \tab religious organizations \tab residential care facilities, without nursing \tab resin, synthetic rubber and fibers, and filaments manufacturing \tab restaurants and other food services \tab rubber product, except tire, manufacturing \tab sawmills and wood preservation \tab scientific research and development services \tab seafood and other miscellaneous foods, n.e.c. \tab securities, commodities, funds, trusts, and other financial investments \tab services incidental to transportation \tab services to buildings and dwellings \tab sewage treatment facilities \tab sewing, needlework, and piece goods stores \tab soap, cleaning compound, and cosmetics manufacturing \tab specialty food stores \tab sporting and athletic goods, and doll, toy and game manufacturing \tab structural metals, and boiler, tank, and shipping container manufacturing \tab sugar and confectionery products \tab taxi and limousine service \tab textile product mills, except carpet and rug \tab tire manufacturing \tab tobacco manufacturing \tab traveler accommodation \tab truck transportation \tab u.s. air force \tab uncodeable \tab veneer, plywood, and engineered wood products \tab warehousing and storage \tab waste management and remediation services \tab water transportation \tab water, steam, air-conditioning, and irrigation systems \tab wholesale electronic markets, agents and brokers \tab wired telecommunications carriers \tab active duty military personnel, all other \tab administration of economic programs and space research \tab administration of human resource programs \tab aluminum production and processing \tab animal food, grain, and oilseed milling \tab apparel accessories and other apparel manufacturing \tab bakeries, except retail \tab beer, wine, and liquor stores \tab coal mining \tab coating, engraving, heat treating and allied activities \tab commercial and service industry machinery manufacturing \tab community food and housing, and emergency services \tab couriers and messengers \tab cutlery and hand tool manufacturing \tab drinking places, alcoholic beverages \tab electronic and precision equipment repair and maintenance \tab engines, turbines, and power transmission equipment manufacturing \tab footwear and leather goods repair \tab foundries \tab funeral homes, and cemeteries and crematories \tab hardware stores \tab investigation and security services \tab knitting fabric mills, and apparel knitting mills \tab labor unions \tab landscaping services \tab lumber and other construction materials, merchant wholesalers \tab mail order houses \tab miscellaneous nondurable goods, merchant wholesalers \tab motion pictures and video industries \tab motor vehicles, parts and supplies, merchant wholesalers \tab museums, art galleries, historical sites, and similar institutions \tab navigational, measuring, electromedical, and control instruments manufacturing \tab not specified retail trade \tab not specified type of mining \tab oil and gas extraction \tab other health care services \tab outpatient care centers \tab paint, coating, and adhesive manufacturing \tab plastics product manufacturing \tab recreational vehicle parks and camps, and rooming and boarding houses \tab retail bakeries \tab ship and boat building \tab specialized design services \tab structural clay product manufacturing \tab support activities for mining \tab u.s. armed forces/n.s. \tab u.s. navy \tab vending machine operators \tab administration of environmental quality and housing programs \tab carpet and rug mills \tab commercial and industrial machinery and equipment repair and maintenance \tab computer systems design and related services \tab logging \tab metal forgings and stampings \tab miscellaneous durable goods, merchant wholesalers \tab nonmetallic mineral mining and quarrying \tab not specified utilities \tab other telecommunications services \tab prefabricated wood buildings and mobile homes \tab savings institutions, including credit unions \tab textile and fabric finishing and coating mills \tab u.s. army \tab used merchandise stores \tab alcoholic beverages, merchant wholesalers \tab automotive equipment rental and leasing \tab data processing, hosting, and related services \tab fishing, hunting, and trapping \tab furniture and home furnishing, merchant wholesalers \tab nail salons and other personal care services \tab other transportation equipment manufacturing \tab shoe stores \tab support activities for agriculture and forestry \tab veterinary services \tab vocational rehabilitation services \tab book stores and news dealers \tab farm supplies, merchant wholesalers \tab miscellaneous petroleum and coal products \tab music stores \tab retail florists \tab u.s. marine corps \tab offices of chiropractors \tab recyclable material, merchant wholesalers \tab travel arrangements and reservation services \tab home health care services \tab offices of other health practitioners \tab other consumer goods rental \tab pipeline transportation \tab scenic and sightseeing transportation \tab gift, novelty, and souvenir shops \tab sporting goods, camera, and hobby and toy stores \tab commercial, industrial, and other intangible assets rental and leasing \tab sound recording industries \tab u.s. coast guard \tab car washes \tab video tape and disk rental \tab software publishers \tab electronic shopping \tab electronic auctions \tab internet publishing and broadcasting and web search portals \tab other information services \tab skipped on web \tab Total \cr
   1972 \tab 4 \tab 3 \tab 3 \tab 1 \tab 3 \tab 4 \tab 8 \tab 6 \tab 3 \tab 3 \tab 5 \tab 2 \tab 7 \tab 5 \tab 15 \tab 1 \tab 16 \tab 4 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 13 \tab 9 \tab 5 \tab 1 \tab 61 \tab 3 \tab 5 \tab 1 \tab 2 \tab 20 \tab 2 \tab 5 \tab 4 \tab 5 \tab 2 \tab 5 \tab 5 \tab 79 \tab 2 \tab 3 \tab 3 \tab 4 \tab 3 \tab 4 \tab 1 \tab 2 \tab 1 \tab 8 \tab 4 \tab 9 \tab 1 \tab 9 \tab 9 \tab 4 \tab 1 \tab 36 \tab 1 \tab 1 \tab 580 \tab 1 \tab 6 \tab 10 \tab 26 \tab 15 \tab 2 \tab 20 \tab 1 \tab 1 \tab 5 \tab 1 \tab 2 \tab 7 \tab 5 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 3 \tab 3 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 14 \tab 23 \tab 1 \tab 3 \tab 33 \tab 5 \tab 6 \tab 2 \tab 1 \tab 6 \tab 1 \tab 1 \tab 6 \tab 2 \tab 2 \tab 1 \tab 12 \tab 2 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 4 \tab 1 \tab 4 \tab 13 \tab 2 \tab 2 \tab 16 \tab 1 \tab 1 \tab 9 \tab 2 \tab 1 \tab 15 \tab 1 \tab 13 \tab 3 \tab 1 \tab 3 \tab 16 \tab 4 \tab 2 \tab 3 \tab 1 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 5 \tab 2 \tab 3 \tab 4 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 22 \tab 2 \tab 118 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 14 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 2 \tab 1 \tab 2 \tab 2 \tab 5 \tab 4 \tab 8 \tab 5 \tab 3 \tab 2 \tab 4 \tab - \tab 12 \tab 7 \tab 11 \tab 3 \tab 6 \tab 2 \tab - \tab 7 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 4 \tab 2 \tab 6 \tab 10 \tab 5 \tab 3 \tab 59 \tab 3 \tab 15 \tab 22 \tab - \tab 15 \tab - \tab 7 \tab 1 \tab 9 \tab 2 \tab 6 \tab 4 \tab 65 \tab 2 \tab 6 \tab 5 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab 3 \tab 2 \tab 7 \tab 5 \tab 5 \tab 7 \tab 22 \tab 2 \tab - \tab 33 \tab - \tab 2 \tab 536 \tab 1 \tab 4 \tab 8 \tab 25 \tab 10 \tab 3 \tab 11 \tab 1 \tab - \tab 9 \tab 1 \tab - \tab 5 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 4 \tab 4 \tab 3 \tab 5 \tab - \tab 2 \tab 2 \tab 1 \tab 20 \tab 19 \tab 3 \tab 4 \tab 30 \tab 5 \tab 2 \tab - \tab 2 \tab 8 \tab - \tab - \tab 6 \tab - \tab 3 \tab - \tab 9 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab 4 \tab 1 \tab 4 \tab 8 \tab - \tab - \tab 1 \tab 1 \tab - \tab 10 \tab - \tab 9 \tab 2 \tab 2 \tab 1 \tab 23 \tab 3 \tab 4 \tab 2 \tab 1 \tab 2 \tab 3 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab 7 \tab 12 \tab 5 \tab 61 \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab 11 \tab 2 \tab 3 \tab 5 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 4 \tab 5 \tab 1 \tab - \tab 4 \tab 2 \tab 5 \tab 6 \tab 12 \tab 2 \tab 4 \tab 3 \tab 1 \tab 3 \tab 5 \tab 2 \tab - \tab - \tab 3 \tab 3 \tab 2 \tab 9 \tab 7 \tab 4 \tab 4 \tab 60 \tab 5 \tab 10 \tab 14 \tab 1 \tab 11 \tab 1 \tab 7 \tab 1 \tab 4 \tab 2 \tab 14 \tab 2 \tab 68 \tab 3 \tab 6 \tab 4 \tab - \tab 1 \tab 4 \tab 1 \tab 4 \tab 1 \tab 7 \tab 4 \tab 5 \tab 4 \tab 4 \tab 21 \tab 1 \tab - \tab 35 \tab 1 \tab 1 \tab 523 \tab 1 \tab 1 \tab 10 \tab 21 \tab 11 \tab 1 \tab 15 \tab 1 \tab - \tab 3 \tab 2 \tab 5 \tab 10 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 7 \tab 6 \tab 12 \tab - \tab 1 \tab 5 \tab 1 \tab 31 \tab 12 \tab 2 \tab 7 \tab 44 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 8 \tab - \tab 3 \tab - \tab 9 \tab - \tab - \tab 7 \tab 2 \tab - \tab 3 \tab 7 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab 3 \tab 2 \tab 3 \tab 6 \tab 4 \tab 12 \tab 3 \tab 3 \tab 12 \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 13 \tab - \tab 5 \tab 6 \tab 1 \tab 3 \tab 17 \tab 1 \tab 2 \tab 2 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab 4 \tab - \tab 2 \tab 4 \tab 9 \tab 3 \tab 26 \tab - \tab 2 \tab - \tab - \tab 4 \tab - \tab 10 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 9 \tab 1 \tab 5 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 5 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 5 \tab 4 \tab 1 \tab 1 \tab 2 \tab 4 \tab 9 \tab 10 \tab 7 \tab 2 \tab 4 \tab 2 \tab 11 \tab 6 \tab 12 \tab 1 \tab 5 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab 2 \tab 10 \tab 7 \tab 2 \tab 70 \tab 5 \tab 13 \tab 7 \tab 1 \tab 30 \tab 4 \tab 5 \tab - \tab 7 \tab - \tab 10 \tab 3 \tab 58 \tab - \tab 3 \tab 3 \tab 2 \tab - \tab 3 \tab - \tab 6 \tab 1 \tab 3 \tab 1 \tab 5 \tab 3 \tab 3 \tab 24 \tab - \tab - \tab 39 \tab 3 \tab 1 \tab 571 \tab 2 \tab 2 \tab 4 \tab 27 \tab 8 \tab 3 \tab 11 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab 6 \tab 4 \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab 4 \tab 3 \tab 5 \tab 4 \tab 2 \tab 2 \tab 2 \tab - \tab 21 \tab 16 \tab 1 \tab 4 \tab 14 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 10 \tab 2 \tab 3 \tab - \tab 11 \tab 3 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab 4 \tab - \tab - \tab 2 \tab 4 \tab 7 \tab 1 \tab 2 \tab 9 \tab 1 \tab 4 \tab 2 \tab - \tab - \tab 10 \tab 1 \tab 11 \tab 2 \tab - \tab 3 \tab 13 \tab 1 \tab 6 \tab 4 \tab 1 \tab 1 \tab 4 \tab 4 \tab - \tab - \tab - \tab 2 \tab 2 \tab 4 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab 4 \tab 11 \tab 3 \tab 46 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 17 \tab 1 \tab 2 \tab 5 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 8 \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 6 \tab 3 \tab 3 \tab - \tab 1 \tab 3 \tab 9 \tab 18 \tab 3 \tab 3 \tab 4 \tab 3 \tab 3 \tab 8 \tab 18 \tab - \tab 6 \tab 2 \tab 1 \tab 3 \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 10 \tab 13 \tab 5 \tab 4 \tab 48 \tab 3 \tab 14 \tab 14 \tab 1 \tab 23 \tab 2 \tab 8 \tab 1 \tab 3 \tab 1 \tab 8 \tab 5 \tab 53 \tab 2 \tab 3 \tab 5 \tab 2 \tab - \tab - \tab 2 \tab 3 \tab - \tab 6 \tab 1 \tab 4 \tab 3 \tab 5 \tab 20 \tab 1 \tab - \tab 29 \tab - \tab 1 \tab 612 \tab 4 \tab 1 \tab 6 \tab 15 \tab 8 \tab 1 \tab 13 \tab 1 \tab 1 \tab 5 \tab 2 \tab 3 \tab 10 \tab 1 \tab 2 \tab 1 \tab 4 \tab 6 \tab - \tab 1 \tab 4 \tab 1 \tab 3 \tab 2 \tab 1 \tab 3 \tab 1 \tab 20 \tab 15 \tab 1 \tab 3 \tab 10 \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 5 \tab - \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 2 \tab 5 \tab 5 \tab - \tab 6 \tab 11 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 15 \tab - \tab 10 \tab 5 \tab 2 \tab 1 \tab 23 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 6 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 6 \tab 2 \tab 30 \tab - \tab 4 \tab 2 \tab - \tab 4 \tab - \tab 12 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 10 \tab 2 \tab 3 \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 4 \tab 1 \tab - \tab 2 \tab 1 \tab 4 \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 6 \tab 6 \tab 6 \tab - \tab 10 \tab 1 \tab 5 \tab 4 \tab 15 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 4 \tab 7 \tab 13 \tab 5 \tab 3 \tab 64 \tab 2 \tab 9 \tab 10 \tab 3 \tab 19 \tab 3 \tab 4 \tab 2 \tab 4 \tab 1 \tab 8 \tab - \tab 51 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 3 \tab 8 \tab 12 \tab 6 \tab 1 \tab 5 \tab 6 \tab 5 \tab - \tab 36 \tab 1 \tab 1 \tab 638 \tab - \tab 4 \tab 9 \tab 18 \tab 8 \tab 4 \tab 11 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 7 \tab 3 \tab - \tab 2 \tab 5 \tab 4 \tab 1 \tab 1 \tab 3 \tab 4 \tab 6 \tab - \tab 2 \tab 3 \tab 3 \tab 22 \tab 14 \tab - \tab 2 \tab 10 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 8 \tab 1 \tab 9 \tab - \tab 3 \tab 1 \tab - \tab 5 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab 3 \tab 3 \tab 3 \tab 4 \tab 1 \tab 5 \tab 9 \tab 1 \tab 4 \tab 12 \tab 1 \tab 2 \tab 2 \tab 3 \tab - \tab 6 \tab 1 \tab 7 \tab 3 \tab 4 \tab 3 \tab 13 \tab 2 \tab 2 \tab 4 \tab - \tab 5 \tab 3 \tab 3 \tab - \tab - \tab 2 \tab - \tab 2 \tab 4 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab 10 \tab 15 \tab 2 \tab 61 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 11 \tab - \tab - \tab 2 \tab 5 \tab 3 \tab - \tab 1 \tab 3 \tab 9 \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 6 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 3 \tab 4 \tab 1 \tab - \tab 1 \tab 3 \tab 7 \tab 8 \tab 2 \tab - \tab 4 \tab 4 \tab 12 \tab 7 \tab 8 \tab 1 \tab 6 \tab 1 \tab 1 \tab 4 \tab 6 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 4 \tab 10 \tab 2 \tab 2 \tab 68 \tab 2 \tab 13 \tab 20 \tab - \tab 20 \tab 2 \tab 3 \tab - \tab 5 \tab - \tab 10 \tab 4 \tab 57 \tab 3 \tab 11 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 3 \tab 4 \tab 11 \tab 7 \tab 3 \tab 4 \tab 14 \tab - \tab - \tab 31 \tab - \tab 1 \tab 653 \tab 3 \tab 2 \tab 5 \tab 24 \tab 13 \tab - \tab 9 \tab 3 \tab - \tab 7 \tab 1 \tab 3 \tab 4 \tab 1 \tab - \tab 3 \tab 3 \tab 3 \tab - \tab 3 \tab 5 \tab 4 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 19 \tab 9 \tab 1 \tab 2 \tab 11 \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 5 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 4 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 3 \tab 3 \tab 3 \tab 1 \tab 1 \tab 5 \tab 4 \tab 1 \tab 12 \tab 7 \tab 4 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 8 \tab 3 \tab 1 \tab 3 \tab 23 \tab 1 \tab 4 \tab 2 \tab - \tab 3 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 7 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 5 \tab 6 \tab 4 \tab - \tab 1 \tab 3 \tab 4 \tab 2 \tab - \tab - \tab 7 \tab - \tab 4 \tab 2 \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab 11 \tab - \tab 5 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 6 \tab - \tab 16 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 5 \tab - \tab 1 \tab - \tab - \tab 4 \tab 5 \tab 7 \tab 3 \tab 2 \tab 7 \tab 2 \tab 13 \tab 9 \tab 13 \tab - \tab 7 \tab - \tab - \tab 6 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab 4 \tab 3 \tab 10 \tab 1 \tab 4 \tab 55 \tab 2 \tab 11 \tab 9 \tab 2 \tab 15 \tab - \tab 4 \tab 4 \tab 13 \tab 2 \tab 8 \tab 4 \tab 52 \tab 1 \tab 5 \tab 8 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 5 \tab 9 \tab 2 \tab 1 \tab 34 \tab 1 \tab - \tab 640 \tab 3 \tab 4 \tab 11 \tab 17 \tab 10 \tab 2 \tab 12 \tab - \tab - \tab 7 \tab 3 \tab 5 \tab 13 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab 4 \tab 2 \tab 13 \tab 14 \tab - \tab 3 \tab 7 \tab 1 \tab 1 \tab 3 \tab - \tab 5 \tab 1 \tab - \tab 8 \tab 1 \tab 3 \tab 1 \tab 7 \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab 4 \tab 4 \tab 1 \tab 5 \tab 4 \tab 3 \tab 3 \tab - \tab 2 \tab - \tab 6 \tab 3 \tab 8 \tab 6 \tab - \tab 1 \tab 25 \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 2 \tab 4 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab 6 \tab 12 \tab 3 \tab 15 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 10 \tab - \tab 3 \tab 6 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 13 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 6 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 9 \tab 8 \tab 3 \tab 1 \tab 4 \tab 4 \tab 6 \tab 9 \tab 12 \tab 1 \tab 10 \tab 2 \tab 1 \tab 2 \tab 7 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 5 \tab 7 \tab 13 \tab 3 \tab 4 \tab 60 \tab 1 \tab 12 \tab 18 \tab 3 \tab 14 \tab 1 \tab 6 \tab 1 \tab 8 \tab - \tab 11 \tab 2 \tab 62 \tab 1 \tab 6 \tab 6 \tab 1 \tab 2 \tab 4 \tab 1 \tab 3 \tab 1 \tab 2 \tab 5 \tab 6 \tab 2 \tab 3 \tab 29 \tab 1 \tab 2 \tab 42 \tab 1 \tab 1 \tab 939 \tab 2 \tab 4 \tab 8 \tab 26 \tab 8 \tab 1 \tab 15 \tab 2 \tab 2 \tab 5 \tab 3 \tab - \tab 6 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 5 \tab 1 \tab 2 \tab 3 \tab - \tab 2 \tab 3 \tab 1 \tab 14 \tab 6 \tab 1 \tab 3 \tab 23 \tab 6 \tab 3 \tab 3 \tab - \tab 3 \tab 1 \tab 2 \tab 10 \tab - \tab 4 \tab - \tab 4 \tab 1 \tab - \tab 4 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 4 \tab 4 \tab - \tab - \tab 4 \tab 3 \tab 17 \tab 2 \tab 6 \tab 7 \tab 1 \tab 3 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab 11 \tab 4 \tab 3 \tab - \tab 22 \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab 1 \tab 3 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab 5 \tab 10 \tab 2 \tab 7 \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab 5 \tab - \tab 3 \tab 3 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 9 \tab 1 \tab 6 \tab - \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 3 \tab 4 \tab 2 \tab - \tab 1 \tab 1 \tab 8 \tab 8 \tab 9 \tab 1 \tab 2 \tab 4 \tab 4 \tab 9 \tab 18 \tab 1 \tab 8 \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 2 \tab 7 \tab 20 \tab 2 \tab 6 \tab 49 \tab 6 \tab 15 \tab 17 \tab 1 \tab 19 \tab 1 \tab 3 \tab 1 \tab 8 \tab 1 \tab 3 \tab 4 \tab 43 \tab 1 \tab 6 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 10 \tab 2 \tab 2 \tab 6 \tab 15 \tab - \tab 1 \tab 40 \tab 1 \tab 2 \tab 704 \tab 2 \tab 4 \tab 8 \tab 26 \tab 7 \tab - \tab 20 \tab 1 \tab 3 \tab 6 \tab 4 \tab 1 \tab 5 \tab 4 \tab 3 \tab 1 \tab 4 \tab 4 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 10 \tab 14 \tab 1 \tab 3 \tab 16 \tab 3 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 14 \tab - \tab 6 \tab 1 \tab 9 \tab - \tab 1 \tab 6 \tab 3 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 8 \tab 1 \tab 3 \tab 9 \tab 1 \tab 2 \tab 2 \tab 1 \tab 4 \tab 5 \tab - \tab 18 \tab 8 \tab 4 \tab 3 \tab 17 \tab - \tab - \tab - \tab 1 \tab 6 \tab 2 \tab 3 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 4 \tab 3 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 16 \tab - \tab 11 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab - \tab 13 \tab - \tab 4 \tab 6 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 5 \tab 1 \tab 5 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab 4 \tab 1 \tab - \tab - \tab 3 \tab - \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1 \tab - \tab 3 \tab - \tab 2 \tab 3 \tab 6 \tab 9 \tab 4 \tab 1 \tab 4 \tab 2 \tab 6 \tab 11 \tab 13 \tab 1 \tab 4 \tab - \tab - \tab 6 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 3 \tab 3 \tab 12 \tab - \tab 2 \tab 57 \tab 2 \tab 11 \tab 11 \tab 3 \tab 15 \tab - \tab 2 \tab 3 \tab 11 \tab - \tab 10 \tab 8 \tab 44 \tab 1 \tab 1 \tab 7 \tab 2 \tab - \tab 3 \tab - \tab 4 \tab 1 \tab 5 \tab 2 \tab 6 \tab 4 \tab 7 \tab 11 \tab 1 \tab 2 \tab 39 \tab 1 \tab 1 \tab 687 \tab 3 \tab 2 \tab 6 \tab 13 \tab 12 \tab 3 \tab 11 \tab 3 \tab - \tab 6 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 4 \tab 4 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 18 \tab 9 \tab 2 \tab 3 \tab 9 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 2 \tab 7 \tab 1 \tab 3 \tab 1 \tab 4 \tab - \tab 1 \tab 3 \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 3 \tab 4 \tab 2 \tab 6 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 6 \tab - \tab 6 \tab 3 \tab 3 \tab 1 \tab 21 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab 5 \tab 1 \tab 1 \tab 2 \tab 4 \tab 2 \tab 5 \tab - \tab 3 \tab - \tab - \tab - \tab 7 \tab 7 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 15 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 4 \tab - \tab 6 \tab - \tab 2 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab 3 \tab 1 \tab 3 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 2 \tab 3 \tab - \tab 6 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 6 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 4 \tab - \tab 3 \tab - \tab - \tab 4 \tab 4 \tab 10 \tab 9 \tab 2 \tab 5 \tab 5 \tab 5 \tab 2 \tab 24 \tab 2 \tab 7 \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab - \tab 1 \tab 8 \tab 1 \tab 9 \tab 11 \tab 3 \tab 1 \tab 45 \tab 1 \tab 12 \tab 12 \tab 1 \tab 17 \tab 1 \tab 5 \tab 3 \tab 8 \tab - \tab 1 \tab 6 \tab 52 \tab 2 \tab 9 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 6 \tab 7 \tab 3 \tab 1 \tab 7 \tab 13 \tab 5 \tab 1 \tab 29 \tab 3 \tab 2 \tab 709 \tab 2 \tab 2 \tab 10 \tab 15 \tab 15 \tab 2 \tab 8 \tab 1 \tab - \tab 5 \tab 2 \tab 5 \tab 7 \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab - \tab - \tab 3 \tab 2 \tab 12 \tab 10 \tab - \tab 3 \tab 8 \tab 5 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 7 \tab 1 \tab 5 \tab - \tab 7 \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 6 \tab 2 \tab 3 \tab - \tab 5 \tab 2 \tab 3 \tab 10 \tab 1 \tab 7 \tab 9 \tab - \tab 3 \tab 4 \tab 4 \tab 3 \tab 7 \tab - \tab 5 \tab 3 \tab 3 \tab - \tab 27 \tab 1 \tab - \tab - \tab 1 \tab 6 \tab 4 \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab 3 \tab 2 \tab - \tab - \tab 1 \tab - \tab 3 \tab 13 \tab - \tab 23 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 2 \tab 7 \tab - \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 5 \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 4 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 5 \tab 1 \tab 5 \tab - \tab - \tab 3 \tab 2 \tab 9 \tab 5 \tab - \tab 7 \tab 4 \tab 3 \tab 3 \tab 12 \tab 1 \tab 6 \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 1 \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 13 \tab 1 \tab - \tab 50 \tab 1 \tab 10 \tab 9 \tab - \tab 14 \tab - \tab 5 \tab 1 \tab 5 \tab - \tab 5 \tab 3 \tab 60 \tab 3 \tab 3 \tab 6 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 5 \tab 7 \tab 2 \tab - \tab 3 \tab 10 \tab 14 \tab 1 \tab - \tab 34 \tab - \tab 2 \tab 698 \tab 4 \tab 5 \tab 4 \tab 14 \tab 12 \tab 2 \tab 14 \tab 3 \tab - \tab 9 \tab 2 \tab 1 \tab 5 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 4 \tab 1 \tab 1 \tab 2 \tab 3 \tab 14 \tab 14 \tab 1 \tab 7 \tab 10 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 7 \tab 2 \tab 2 \tab - \tab 6 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 7 \tab 1 \tab 4 \tab 6 \tab 4 \tab 2 \tab 1 \tab 3 \tab 3 \tab 3 \tab - \tab 8 \tab 7 \tab 3 \tab 1 \tab 20 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 7 \tab 14 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 7 \tab - \tab 4 \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 3 \tab 5 \tab 3 \tab 2 \tab - \tab 4 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab 2 \tab 3 \tab 11 \tab 5 \tab - \tab 1 \tab 6 \tab 10 \tab 6 \tab 17 \tab 1 \tab 5 \tab 1 \tab - \tab 5 \tab 4 \tab 2 \tab - \tab 1 \tab 1 \tab 5 \tab 2 \tab 6 \tab 9 \tab 1 \tab 5 \tab 56 \tab 1 \tab 13 \tab 8 \tab - \tab 15 \tab 1 \tab 1 \tab 3 \tab 7 \tab 1 \tab 4 \tab 5 \tab 55 \tab - \tab 6 \tab 3 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 5 \tab 3 \tab 4 \tab 3 \tab 10 \tab 22 \tab 1 \tab - \tab 29 \tab 2 \tab 1 \tab 949 \tab 2 \tab 5 \tab 7 \tab 17 \tab 6 \tab 3 \tab 10 \tab 3 \tab - \tab 6 \tab 2 \tab 2 \tab 4 \tab 2 \tab 1 \tab 4 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 6 \tab 1 \tab 3 \tab 5 \tab 2 \tab 19 \tab 21 \tab 1 \tab 6 \tab 17 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 11 \tab 1 \tab 7 \tab - \tab 5 \tab 1 \tab 2 \tab 13 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 5 \tab 7 \tab - \tab 5 \tab 7 \tab 4 \tab 1 \tab - \tab 1 \tab 4 \tab 6 \tab - \tab 10 \tab 2 \tab 3 \tab 2 \tab 17 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 6 \tab 13 \tab 2 \tab 24 \tab - \tab 3 \tab 2 \tab 1 \tab 4 \tab - \tab 7 \tab - \tab 6 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 7 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 5 \tab - \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 6 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 5 \tab - \tab 5 \tab - \tab 2 \tab 4 \tab 10 \tab 8 \tab - \tab 3 \tab 1 \tab 2 \tab 4 \tab 7 \tab 11 \tab 2 \tab 9 \tab 2 \tab - \tab 5 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 10 \tab 4 \tab 4 \tab 13 \tab 1 \tab 3 \tab 49 \tab 1 \tab 4 \tab 6 \tab 1 \tab 19 \tab - \tab 5 \tab - \tab 12 \tab 2 \tab 2 \tab 4 \tab 51 \tab 3 \tab 11 \tab 3 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 4 \tab 2 \tab 6 \tab 6 \tab - \tab 1 \tab 27 \tab - \tab 1 \tab 742 \tab 3 \tab 2 \tab 8 \tab 19 \tab 6 \tab - \tab 9 \tab - \tab - \tab 8 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 21 \tab 3 \tab 1 \tab 3 \tab 13 \tab 3 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab 8 \tab 2 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 6 \tab 1 \tab 3 \tab 7 \tab 1 \tab 3 \tab 1 \tab 4 \tab 2 \tab 4 \tab - \tab 16 \tab 3 \tab 3 \tab - \tab 22 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 2 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 11 \tab 2 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 2 \tab 6 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 3 \tab 2 \tab 4 \tab - \tab 3 \tab 2 \tab 4 \tab 11 \tab 5 \tab 1 \tab 5 \tab 1 \tab 6 \tab 7 \tab 14 \tab 1 \tab 2 \tab 1 \tab - \tab 7 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 6 \tab 1 \tab 8 \tab 17 \tab - \tab 2 \tab 53 \tab 2 \tab 10 \tab 14 \tab 1 \tab 11 \tab 2 \tab 3 \tab - \tab 4 \tab - \tab 4 \tab 2 \tab 39 \tab 2 \tab 9 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 4 \tab 4 \tab 3 \tab 7 \tab 18 \tab 1 \tab 1 \tab 36 \tab - \tab 2 \tab 734 \tab 2 \tab 2 \tab 5 \tab 20 \tab 11 \tab - \tab 10 \tab 1 \tab 1 \tab 10 \tab - \tab 2 \tab 6 \tab 3 \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab 15 \tab 17 \tab 2 \tab 7 \tab 18 \tab 1 \tab 2 \tab 2 \tab 1 \tab 7 \tab 2 \tab - \tab 10 \tab 1 \tab 4 \tab - \tab 8 \tab - \tab 3 \tab 7 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 5 \tab 2 \tab 9 \tab 5 \tab - \tab 4 \tab 3 \tab 2 \tab - \tab 7 \tab 1 \tab 13 \tab 8 \tab 4 \tab - \tab 20 \tab 2 \tab 3 \tab 2 \tab - \tab 4 \tab 2 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 10 \tab - \tab 2 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab 3 \tab 8 \tab 1 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 5 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 6 \tab 2 \tab 3 \tab - \tab - \tab 5 \tab 5 \tab 8 \tab 2 \tab - \tab 6 \tab 3 \tab 7 \tab - \tab 11 \tab - \tab 8 \tab 1 \tab - \tab 3 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 13 \tab 2 \tab 1 \tab 18 \tab - \tab 3 \tab 55 \tab 3 \tab 8 \tab 11 \tab - \tab 14 \tab 1 \tab 2 \tab - \tab 6 \tab 2 \tab 4 \tab 5 \tab 40 \tab 2 \tab 6 \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 4 \tab 2 \tab 5 \tab 12 \tab - \tab - \tab 21 \tab - \tab - \tab 684 \tab 1 \tab 4 \tab 3 \tab 19 \tab 5 \tab - \tab 10 \tab - \tab 1 \tab 9 \tab - \tab - \tab 6 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 3 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 12 \tab 8 \tab 1 \tab 2 \tab 17 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 5 \tab - \tab 5 \tab - \tab 6 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab - \tab 6 \tab 4 \tab 1 \tab - \tab 4 \tab 3 \tab 2 \tab 7 \tab - \tab 1 \tab - \tab 4 \tab - \tab 17 \tab 9 \tab 2 \tab - \tab 20 \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 3 \tab - \tab - \tab 3 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 6 \tab - \tab 3 \tab 7 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 4 \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 5 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 4 \tab - \tab 1 \tab - \tab 2 \tab - \tab 4 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 6 \tab 2 \tab 6 \tab - \tab - \tab 1 \tab 4 \tab 6 \tab 3 \tab 1 \tab 3 \tab 3 \tab 6 \tab 13 \tab 10 \tab - \tab 5 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab 3 \tab 14 \tab 1 \tab 3 \tab 52 \tab 2 \tab 5 \tab 6 \tab 2 \tab 18 \tab - \tab 2 \tab 2 \tab 10 \tab 2 \tab 2 \tab 2 \tab 52 \tab 1 \tab 5 \tab 3 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 1 \tab 4 \tab 19 \tab 1 \tab - \tab 31 \tab - \tab - \tab 748 \tab 7 \tab 4 \tab 3 \tab 17 \tab 9 \tab 4 \tab 16 \tab 3 \tab - \tab 8 \tab - \tab 2 \tab 6 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 5 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 16 \tab 7 \tab - \tab 2 \tab 17 \tab 3 \tab - \tab - \tab - \tab 6 \tab - \tab 1 \tab 12 \tab 3 \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab 4 \tab - \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 5 \tab - \tab 5 \tab 9 \tab - \tab 4 \tab 8 \tab 1 \tab 4 \tab - \tab 3 \tab - \tab 2 \tab - \tab 9 \tab 7 \tab 3 \tab 2 \tab 25 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 7 \tab 14 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab - \tab 4 \tab 2 \tab 7 \tab - \tab 4 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 6 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 6 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 2 \tab 8 \tab 1 \tab - \tab 2 \tab 4 \tab 8 \tab 7 \tab 2 \tab 2 \tab 6 \tab 1 \tab 3 \tab 2 \tab 19 \tab 1 \tab 5 \tab 3 \tab - \tab 2 \tab 3 \tab 3 \tab - \tab 1 \tab 1 \tab 6 \tab 3 \tab 7 \tab 14 \tab 3 \tab 3 \tab 65 \tab 2 \tab 9 \tab 16 \tab - \tab 10 \tab - \tab 1 \tab 1 \tab 6 \tab - \tab 4 \tab 3 \tab 42 \tab 2 \tab 10 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 4 \tab 4 \tab 4 \tab 2 \tab 4 \tab 17 \tab 2 \tab 2 \tab 41 \tab 1 \tab - \tab 781 \tab 5 \tab 6 \tab 4 \tab 18 \tab 4 \tab 1 \tab 19 \tab 2 \tab - \tab 17 \tab 3 \tab - \tab 4 \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 4 \tab 3 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 12 \tab 9 \tab 2 \tab 3 \tab 15 \tab - \tab - \tab 2 \tab - \tab 8 \tab - \tab 3 \tab 7 \tab - \tab 1 \tab - \tab 8 \tab 2 \tab 2 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 8 \tab 2 \tab 8 \tab 8 \tab 2 \tab 4 \tab - \tab 2 \tab 3 \tab 4 \tab - \tab 12 \tab 6 \tab 1 \tab - \tab 22 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 5 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 7 \tab 11 \tab 4 \tab 2 \tab - \tab - \tab 3 \tab 2 \tab - \tab 2 \tab 8 \tab - \tab 9 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 3 \tab - \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 4 \tab 4 \tab 10 \tab 1 \tab - \tab 4 \tab 9 \tab 14 \tab 4 \tab 3 \tab 14 \tab 4 \tab 6 \tab 10 \tab 25 \tab 1 \tab 6 \tab 4 \tab 4 \tab 10 \tab 5 \tab 3 \tab 1 \tab 1 \tab 3 \tab 14 \tab 6 \tab 11 \tab 29 \tab 2 \tab 9 \tab 104 \tab 2 \tab 6 \tab 15 \tab 7 \tab 23 \tab 3 \tab 9 \tab 1 \tab 12 \tab 5 \tab 3 \tab 5 \tab 103 \tab 5 \tab 8 \tab 7 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 5 \tab 7 \tab 8 \tab - \tab 4 \tab 23 \tab 3 \tab 2 \tab 61 \tab - \tab - \tab 1521 \tab 6 \tab 5 \tab 16 \tab 36 \tab 4 \tab 5 \tab 28 \tab 3 \tab - \tab 10 \tab 1 \tab 3 \tab 7 \tab 4 \tab - \tab 2 \tab 6 \tab - \tab 1 \tab 1 \tab 3 \tab 4 \tab 6 \tab 2 \tab 2 \tab 1 \tab - \tab 29 \tab 21 \tab 5 \tab 5 \tab 26 \tab 11 \tab 1 \tab 1 \tab 2 \tab 5 \tab - \tab 2 \tab 17 \tab 1 \tab 9 \tab 2 \tab 18 \tab - \tab - \tab 14 \tab 2 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab 5 \tab 3 \tab 1 \tab 5 \tab 6 \tab 6 \tab 10 \tab 2 \tab 10 \tab 8 \tab 8 \tab 7 \tab 6 \tab 8 \tab 5 \tab 7 \tab - \tab 29 \tab 4 \tab 6 \tab 5 \tab 41 \tab 3 \tab 1 \tab 4 \tab - \tab 5 \tab 9 \tab 9 \tab - \tab - \tab 3 \tab 2 \tab - \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 17 \tab 16 \tab 8 \tab 2 \tab - \tab 6 \tab 4 \tab - \tab 3 \tab 1 \tab 9 \tab - \tab 5 \tab 16 \tab - \tab 2 \tab - \tab 6 \tab 2 \tab - \tab 1 \tab 9 \tab - \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 4 \tab 1 \tab 1 \tab 7 \tab 2 \tab 1 \tab 1 \tab 5 \tab 3 \tab 3 \tab - \tab 8 \tab - \tab 2 \tab 5 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 3 \tab 9 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 7 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 2 \tab 7 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 13 \tab 3 \tab 5 \tab 1 \tab 1 \tab 6 \tab 10 \tab 10 \tab 2 \tab 3 \tab 13 \tab 4 \tab 9 \tab 9 \tab 18 \tab 2 \tab 7 \tab 3 \tab - \tab 7 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab 14 \tab 3 \tab 9 \tab 18 \tab 3 \tab 4 \tab 94 \tab - \tab 8 \tab 8 \tab 2 \tab 21 \tab 3 \tab 2 \tab - \tab 7 \tab 3 \tab 7 \tab 14 \tab 96 \tab 4 \tab 12 \tab 6 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 9 \tab 7 \tab 4 \tab - \tab 9 \tab 12 \tab 1 \tab 4 \tab 57 \tab 2 \tab 1 \tab 1563 \tab 8 \tab 7 \tab 8 \tab 22 \tab 5 \tab 3 \tab 22 \tab 1 \tab - \tab 16 \tab 3 \tab 1 \tab 10 \tab 1 \tab 1 \tab 6 \tab 2 \tab - \tab - \tab 4 \tab 5 \tab 4 \tab 2 \tab - \tab 5 \tab 4 \tab 2 \tab 22 \tab 17 \tab - \tab 5 \tab 18 \tab 5 \tab - \tab 1 \tab 1 \tab 19 \tab - \tab 4 \tab 14 \tab 1 \tab 6 \tab - \tab 14 \tab 1 \tab 2 \tab 7 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 4 \tab 1 \tab 2 \tab 3 \tab 3 \tab 2 \tab 8 \tab 2 \tab 6 \tab 10 \tab - \tab 8 \tab 8 \tab 7 \tab 2 \tab 7 \tab 4 \tab 9 \tab 10 \tab - \tab 21 \tab 16 \tab 7 \tab 5 \tab 45 \tab 2 \tab 2 \tab 1 \tab - \tab 10 \tab 6 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 2 \tab - \tab 11 \tab 18 \tab 6 \tab 2 \tab - \tab 4 \tab - \tab - \tab 3 \tab 3 \tab 14 \tab 1 \tab 8 \tab 3 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 5 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 7 \tab 1 \tab 1 \tab 1 \tab 14 \tab 2 \tab - \tab 9 \tab 7 \tab 1 \tab 6 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab 3 \tab 15 \tab 1 \tab - \tab - \tab - \tab 2 \tab 5 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab 6 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 4 \tab 6 \tab 7 \tab - \tab 1 \tab 8 \tab 12 \tab 4 \tab 2 \tab 1 \tab 12 \tab 2 \tab 10 \tab 11 \tab 18 \tab - \tab 14 \tab 1 \tab - \tab 4 \tab 5 \tab 1 \tab 1 \tab - \tab 2 \tab 15 \tab 3 \tab 7 \tab 29 \tab 4 \tab 7 \tab 85 \tab 2 \tab 10 \tab 5 \tab 3 \tab 16 \tab 1 \tab 3 \tab 2 \tab 6 \tab 3 \tab 8 \tab 8 \tab 81 \tab 3 \tab 9 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 9 \tab 9 \tab 5 \tab 1 \tab 5 \tab 15 \tab 1 \tab 2 \tab 55 \tab 4 \tab 2 \tab 1527 \tab 3 \tab 10 \tab 8 \tab 21 \tab 7 \tab 6 \tab 30 \tab 5 \tab - \tab 10 \tab 2 \tab 4 \tab 5 \tab 4 \tab - \tab 5 \tab 3 \tab - \tab 1 \tab 4 \tab 1 \tab 5 \tab 3 \tab 2 \tab 3 \tab 1 \tab 3 \tab 27 \tab 20 \tab 4 \tab 3 \tab 35 \tab 4 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 2 \tab 12 \tab - \tab 5 \tab 1 \tab 15 \tab - \tab 2 \tab 9 \tab 3 \tab 1 \tab 3 \tab 1 \tab 8 \tab 3 \tab 1 \tab 1 \tab 4 \tab 1 \tab 3 \tab 3 \tab 6 \tab 4 \tab 10 \tab - \tab 6 \tab 10 \tab 3 \tab 4 \tab 7 \tab 5 \tab 3 \tab 6 \tab 1 \tab 31 \tab 7 \tab 4 \tab 3 \tab 37 \tab 1 \tab 3 \tab 2 \tab 1 \tab 10 \tab 5 \tab 5 \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 3 \tab 2 \tab 3 \tab - \tab - \tab - \tab 5 \tab 25 \tab 6 \tab - \tab - \tab 4 \tab 5 \tab 1 \tab 2 \tab 2 \tab 9 \tab 1 \tab 2 \tab 5 \tab 3 \tab - \tab - \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 3 \tab 6 \tab 1 \tab 2 \tab 4 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab 3 \tab 1 \tab - \tab 9 \tab 8 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 7 \tab 1 \tab 3 \tab 1 \tab 4 \tab - \tab 3 \tab 2 \tab 2 \tab 12 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 5 \tab 1 \tab 8 \tab - \tab 4 \tab 2 \tab 3 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 5 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 5 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 9 \tab 4 \tab 5 \tab 1 \tab - \tab 7 \tab 10 \tab 14 \tab 7 \tab 1 \tab 12 \tab 2 \tab 11 \tab 10 \tab 22 \tab - \tab 5 \tab 2 \tab - \tab 11 \tab 5 \tab 5 \tab 1 \tab 1 \tab 3 \tab 12 \tab 3 \tab 7 \tab 24 \tab 4 \tab 5 \tab 77 \tab - \tab 7 \tab 12 \tab 3 \tab 19 \tab 3 \tab 2 \tab - \tab 6 \tab - \tab 2 \tab 6 \tab 84 \tab 4 \tab 5 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab 6 \tab 4 \tab 1 \tab 5 \tab 11 \tab - \tab 1 \tab 36 \tab 2 \tab 3 \tab 1608 \tab 8 \tab 8 \tab 10 \tab 23 \tab 5 \tab 2 \tab 17 \tab 4 \tab - \tab 17 \tab 3 \tab 3 \tab 4 \tab 2 \tab - \tab 5 \tab 4 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 6 \tab 2 \tab 2 \tab 2 \tab 3 \tab 25 \tab 15 \tab 2 \tab 5 \tab 44 \tab 8 \tab 2 \tab 2 \tab - \tab 4 \tab - \tab 2 \tab 16 \tab 2 \tab 4 \tab 1 \tab 10 \tab 3 \tab 1 \tab 7 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 6 \tab 1 \tab 2 \tab 5 \tab 3 \tab 1 \tab 7 \tab 3 \tab 6 \tab 7 \tab - \tab 6 \tab 7 \tab 2 \tab 2 \tab 7 \tab 1 \tab 5 \tab 6 \tab - \tab 11 \tab 6 \tab 7 \tab 3 \tab 29 \tab 2 \tab 1 \tab 6 \tab 3 \tab 9 \tab 7 \tab 8 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 8 \tab 23 \tab - \tab - \tab - \tab 8 \tab 2 \tab - \tab 1 \tab - \tab 8 \tab - \tab 7 \tab 4 \tab 2 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 4 \tab - \tab 3 \tab - \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 4 \tab - \tab 2 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 4 \tab 6 \tab 2 \tab 4 \tab 3 \tab 3 \tab 3 \tab 5 \tab - \tab 3 \tab - \tab 3 \tab - \tab 4 \tab - \tab 2 \tab 12 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 4 \tab - \tab 2 \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 14 \tab 3 \tab - \tab - \tab - \tab 8 \tab 7 \tab 6 \tab 2 \tab 1 \tab 10 \tab 2 \tab 6 \tab 11 \tab 23 \tab 2 \tab 9 \tab 2 \tab - \tab 7 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 15 \tab 5 \tab 3 \tab 23 \tab 2 \tab 7 \tab 71 \tab 1 \tab 5 \tab 5 \tab 3 \tab 13 \tab 3 \tab 4 \tab - \tab 10 \tab - \tab 5 \tab 7 \tab 94 \tab 2 \tab 8 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 3 \tab 3 \tab 3 \tab 8 \tab 12 \tab - \tab 2 \tab 55 \tab 2 \tab 1 \tab 1547 \tab 3 \tab 9 \tab 12 \tab 19 \tab 11 \tab 4 \tab 26 \tab 1 \tab - \tab 11 \tab 1 \tab 2 \tab 8 \tab 4 \tab 1 \tab 4 \tab 4 \tab - \tab - \tab 4 \tab 5 \tab - \tab 4 \tab 1 \tab 1 \tab 4 \tab - \tab 18 \tab 12 \tab 2 \tab 8 \tab 33 \tab 5 \tab 1 \tab 4 \tab - \tab 10 \tab 1 \tab 2 \tab 11 \tab - \tab 8 \tab 1 \tab 21 \tab - \tab - \tab 7 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 3 \tab 1 \tab 4 \tab 2 \tab 5 \tab 8 \tab - \tab 7 \tab 6 \tab - \tab 9 \tab 3 \tab 5 \tab 6 \tab 3 \tab - \tab 20 \tab 7 \tab - \tab - \tab 38 \tab 1 \tab 1 \tab 2 \tab 2 \tab 9 \tab 9 \tab 6 \tab 2 \tab - \tab 2 \tab 2 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 12 \tab 14 \tab 1 \tab - \tab - \tab 4 \tab 8 \tab - \tab 2 \tab 2 \tab 8 \tab - \tab 5 \tab 7 \tab 1 \tab 4 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 8 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 1 \tab 3 \tab 7 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 4 \tab 10 \tab - \tab 1 \tab 5 \tab 8 \tab - \tab 6 \tab 2 \tab - \tab 1 \tab 4 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab 4 \tab 13 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 2 \tab 4 \tab 8 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 18 \tab 3 \tab 3 \tab - \tab - \tab 4 \tab 5 \tab 4 \tab 7 \tab - \tab 8 \tab 6 \tab 10 \tab 16 \tab 20 \tab 1 \tab 12 \tab - \tab - \tab 7 \tab 5 \tab 6 \tab 2 \tab - \tab 3 \tab 18 \tab 7 \tab 7 \tab 27 \tab 3 \tab 7 \tab 106 \tab 2 \tab 3 \tab 3 \tab - \tab 15 \tab 3 \tab 5 \tab 1 \tab 11 \tab - \tab 5 \tab 8 \tab 105 \tab 4 \tab 13 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 10 \tab 7 \tab 2 \tab 9 \tab 19 \tab - \tab 4 \tab 60 \tab 3 \tab - \tab 1406 \tab 5 \tab 12 \tab 5 \tab 20 \tab 3 \tab 1 \tab 29 \tab 5 \tab 1 \tab 19 \tab 2 \tab 3 \tab 3 \tab 5 \tab 5 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab 4 \tab 8 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab 15 \tab 14 \tab 4 \tab 2 \tab 50 \tab 17 \tab 1 \tab 5 \tab - \tab 12 \tab - \tab 4 \tab 20 \tab 1 \tab 7 \tab 1 \tab 20 \tab 1 \tab 2 \tab 6 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 2 \tab 1 \tab 4 \tab 4 \tab 2 \tab 5 \tab 4 \tab 7 \tab 9 \tab - \tab 9 \tab 8 \tab 3 \tab 8 \tab 4 \tab 3 \tab 5 \tab 5 \tab - \tab 26 \tab 8 \tab 5 \tab 4 \tab 34 \tab 2 \tab 4 \tab 11 \tab - \tab 10 \tab 7 \tab 7 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 11 \tab 28 \tab 3 \tab - \tab - \tab 3 \tab 5 \tab - \tab 5 \tab 2 \tab 16 \tab - \tab 3 \tab 8 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 8 \tab - \tab - \tab 7 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 4 \tab 2 \tab 8 \tab - \tab 1 \tab 1 \tab 5 \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab 2 \tab 2 \tab - \tab 6 \tab 1 \tab 2 \tab - \tab - \tab 8 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab 5 \tab - \tab 3 \tab - \tab 2 \tab 5 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 4 \tab - \tab - \tab 6 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab 4 \tab 3 \tab - \tab 1 \tab - \tab 4 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 12 \tab 9 \tab 8 \tab 1 \tab - \tab 10 \tab 15 \tab 11 \tab 8 \tab - \tab 16 \tab 4 \tab 16 \tab 25 \tab 24 \tab - \tab 11 \tab 6 \tab - \tab 10 \tab 7 \tab 7 \tab - \tab 1 \tab 2 \tab 24 \tab 7 \tab 6 \tab 43 \tab 3 \tab 11 \tab 164 \tab 3 \tab 11 \tab 10 \tab 3 \tab 19 \tab 3 \tab 3 \tab - \tab 10 \tab 3 \tab 6 \tab 8 \tab 152 \tab 4 \tab 29 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 9 \tab 18 \tab 7 \tab 5 \tab 4 \tab 30 \tab 1 \tab 2 \tab 95 \tab 2 \tab 1 \tab 2465 \tab 12 \tab 10 \tab 15 \tab 33 \tab 7 \tab 2 \tab 43 \tab 2 \tab 1 \tab 12 \tab 2 \tab 2 \tab 10 \tab 1 \tab 1 \tab 11 \tab 6 \tab - \tab 1 \tab 4 \tab 1 \tab 5 \tab 5 \tab 3 \tab 4 \tab 5 \tab 2 \tab 31 \tab 19 \tab 3 \tab 6 \tab 44 \tab 10 \tab 1 \tab 3 \tab 3 \tab 26 \tab 3 \tab 1 \tab 19 \tab 4 \tab 14 \tab 1 \tab 22 \tab 1 \tab 5 \tab 9 \tab 3 \tab 2 \tab 2 \tab 2 \tab 2 \tab 7 \tab - \tab 2 \tab 4 \tab 2 \tab 4 \tab 1 \tab 5 \tab 4 \tab 12 \tab - \tab 11 \tab 16 \tab 6 \tab 7 \tab 2 \tab 4 \tab 7 \tab 4 \tab - \tab 35 \tab 19 \tab 7 \tab 8 \tab 62 \tab 1 \tab 2 \tab 8 \tab 5 \tab 21 \tab 10 \tab 11 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 6 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 12 \tab 41 \tab 8 \tab - \tab 2 \tab 7 \tab 6 \tab 1 \tab 2 \tab 1 \tab 19 \tab - \tab 16 \tab 10 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 6 \tab - \tab 9 \tab - \tab 2 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 3 \tab 3 \tab 11 \tab 3 \tab - \tab 3 \tab 7 \tab 1 \tab 8 \tab 5 \tab 20 \tab 1 \tab 1 \tab 8 \tab 10 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 1 \tab 3 \tab 4 \tab 3 \tab 1 \tab - \tab 8 \tab - \tab 8 \tab 20 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 9 \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab 2 \tab 2 \tab 3 \tab - \tab - \tab 4 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab 3 \tab 14 \tab 2 \tab - \tab - \tab 3 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 5 \tab 2 \tab 3 \tab 1 \tab 2 \tab 5 \tab 7 \tab 8 \tab 2 \tab - \tab 11 \tab 3 \tab 10 \tab 4 \tab 18 \tab 1 \tab 1 \tab 1 \tab - \tab 8 \tab 2 \tab 3 \tab - \tab 3 \tab 1 \tab 4 \tab 11 \tab 4 \tab 20 \tab 1 \tab 2 \tab 70 \tab - \tab 7 \tab 4 \tab 2 \tab 12 \tab 1 \tab 1 \tab 1 \tab 7 \tab - \tab 4 \tab 7 \tab 68 \tab 3 \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 6 \tab 7 \tab 3 \tab 2 \tab 6 \tab 9 \tab 1 \tab 1 \tab 38 \tab 1 \tab - \tab 1079 \tab 7 \tab 10 \tab 1 \tab 15 \tab 1 \tab - \tab 25 \tab 3 \tab 1 \tab 7 \tab 4 \tab 3 \tab 2 \tab 3 \tab 1 \tab 4 \tab 4 \tab - \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 8 \tab 8 \tab 2 \tab 2 \tab 19 \tab 8 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 10 \tab - \tab 11 \tab 2 \tab 9 \tab 1 \tab 3 \tab 6 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 4 \tab 1 \tab 3 \tab 5 \tab 1 \tab - \tab 2 \tab 3 \tab 4 \tab 7 \tab - \tab 2 \tab 8 \tab 1 \tab 2 \tab - \tab 3 \tab 2 \tab 4 \tab 1 \tab 11 \tab 9 \tab 8 \tab 2 \tab 20 \tab - \tab 3 \tab 3 \tab - \tab 5 \tab 5 \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab 6 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 15 \tab 10 \tab 1 \tab - \tab 1 \tab 3 \tab 10 \tab - \tab 1 \tab - \tab 10 \tab 1 \tab 9 \tab 10 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 7 \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 3 \tab 2 \tab 3 \tab - \tab - \tab 4 \tab 7 \tab - \tab 3 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 9 \tab 1 \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab 4 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 8 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 3 \tab 4 \tab 2 \tab - \tab 1 \tab 6 \tab 4 \tab 3 \tab 3 \tab - \tab 8 \tab 3 \tab 5 \tab 5 \tab 19 \tab - \tab 3 \tab 1 \tab - \tab 4 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab 8 \tab 9 \tab 3 \tab 22 \tab 2 \tab 9 \tab 57 \tab - \tab 5 \tab 2 \tab 2 \tab 9 \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab 1 \tab 3 \tab 62 \tab 6 \tab 9 \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 6 \tab 5 \tab 2 \tab 1 \tab 12 \tab 3 \tab 2 \tab 44 \tab - \tab - \tab 1187 \tab 2 \tab 3 \tab 2 \tab 15 \tab 6 \tab - \tab 14 \tab - \tab 1 \tab 10 \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 5 \tab 3 \tab - \tab - \tab 2 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab 10 \tab 8 \tab 1 \tab 2 \tab 23 \tab 6 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 7 \tab 1 \tab 9 \tab 1 \tab 13 \tab 1 \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 3 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 4 \tab 1 \tab 4 \tab 8 \tab 1 \tab 6 \tab 8 \tab 3 \tab 3 \tab 2 \tab 3 \tab 4 \tab - \tab - \tab 15 \tab 6 \tab 6 \tab 1 \tab 18 \tab 2 \tab 4 \tab 7 \tab - \tab 6 \tab 4 \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 6 \tab 19 \tab 3 \tab - \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab 9 \tab 1 \tab 6 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 8 \tab 1 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 7 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab 9 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab 6 \tab - \tab 7 \tab 4 \tab 6 \tab 7 \tab 14 \tab - \tab 11 \tab - \tab - \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab 6 \tab 9 \tab 2 \tab 22 \tab 1 \tab 3 \tab 82 \tab 1 \tab 4 \tab 3 \tab 2 \tab 6 \tab - \tab 1 \tab - \tab 12 \tab - \tab 2 \tab 4 \tab 53 \tab 1 \tab 13 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab 2 \tab 2 \tab 8 \tab 6 \tab - \tab - \tab 40 \tab 1 \tab 1 \tab 1121 \tab 3 \tab 3 \tab 3 \tab 15 \tab 2 \tab 2 \tab 16 \tab - \tab 1 \tab 8 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 7 \tab 5 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 13 \tab 1 \tab 1 \tab - \tab 24 \tab 6 \tab - \tab 1 \tab 1 \tab 13 \tab 1 \tab 1 \tab 6 \tab - \tab 3 \tab 1 \tab 9 \tab - \tab 1 \tab 11 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 5 \tab 3 \tab - \tab 3 \tab 8 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 18 \tab 7 \tab 8 \tab 1 \tab 20 \tab - \tab - \tab 6 \tab 2 \tab 5 \tab 4 \tab 5 \tab 1 \tab - \tab 2 \tab 2 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 10 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 6 \tab 1 \tab 4 \tab - \tab - \tab 10 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 6 \tab 1 \tab - \tab 7 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 3 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 34 \tab - \tab - \tab 1 \tab - \tab 1 \tab 10 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 5 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 5 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1974 \cr
   2014 \tab 10 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab 5 \tab 3 \tab 6 \tab 1 \tab 10 \tab 4 \tab 6 \tab 12 \tab 19 \tab 2 \tab 6 \tab 6 \tab - \tab 7 \tab 4 \tab 2 \tab 4 \tab - \tab 2 \tab 8 \tab 10 \tab 4 \tab 22 \tab 1 \tab 2 \tab 73 \tab - \tab 6 \tab 2 \tab 2 \tab 7 \tab 4 \tab 2 \tab 3 \tab 15 \tab 4 \tab 2 \tab 4 \tab 77 \tab 3 \tab 7 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 6 \tab 4 \tab 4 \tab 5 \tab 4 \tab 13 \tab 1 \tab 1 \tab 53 \tab 1 \tab - \tab 1434 \tab 4 \tab 6 \tab 4 \tab 23 \tab 5 \tab - \tab 19 \tab 3 \tab - \tab 16 \tab 3 \tab 2 \tab 9 \tab 2 \tab 1 \tab 9 \tab 7 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 5 \tab 1 \tab - \tab 1 \tab 3 \tab 10 \tab 12 \tab - \tab - \tab 21 \tab 7 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 15 \tab - \tab 6 \tab - \tab 17 \tab - \tab 2 \tab 8 \tab - \tab 1 \tab - \tab - \tab 4 \tab 3 \tab 1 \tab 2 \tab 2 \tab - \tab 3 \tab 1 \tab 4 \tab 6 \tab 9 \tab - \tab 5 \tab 9 \tab 6 \tab 5 \tab 4 \tab 4 \tab 2 \tab 2 \tab 1 \tab 22 \tab 6 \tab 5 \tab 2 \tab 27 \tab 1 \tab 1 \tab 3 \tab 2 \tab 6 \tab 6 \tab 7 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 6 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 19 \tab 6 \tab - \tab - \tab 2 \tab 5 \tab - \tab 3 \tab 2 \tab 5 \tab 1 \tab 5 \tab 6 \tab 2 \tab - \tab - \tab 2 \tab - \tab 2 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 7 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab 10 \tab 2 \tab - \tab 6 \tab 16 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 4 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab 19 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 5 \tab - \tab 3 \tab - \tab 8 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 6 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 6 \tab 10 \tab 2 \tab - \tab - \tab 3 \tab 4 \tab 4 \tab 4 \tab - \tab 6 \tab 4 \tab 8 \tab 14 \tab 16 \tab 2 \tab 7 \tab 3 \tab - \tab 10 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 9 \tab 5 \tab 4 \tab 34 \tab - \tab 2 \tab 96 \tab 2 \tab 7 \tab 1 \tab 3 \tab 12 \tab 2 \tab 1 \tab - \tab 11 \tab - \tab 3 \tab 4 \tab 96 \tab 5 \tab 13 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 5 \tab 2 \tab 5 \tab 14 \tab 2 \tab - \tab 65 \tab - \tab - \tab 1692 \tab 4 \tab 9 \tab 5 \tab 12 \tab 4 \tab 2 \tab 30 \tab 1 \tab - \tab 13 \tab 1 \tab 1 \tab 4 \tab 2 \tab 2 \tab 14 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab 6 \tab 2 \tab - \tab 1 \tab 3 \tab 2 \tab 11 \tab 20 \tab 4 \tab 4 \tab 26 \tab 1 \tab 1 \tab - \tab - \tab 15 \tab 1 \tab 1 \tab 11 \tab - \tab 10 \tab - \tab 9 \tab - \tab 2 \tab 13 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 4 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 4 \tab 6 \tab 5 \tab 1 \tab 4 \tab 7 \tab 3 \tab 4 \tab - \tab 2 \tab 3 \tab 3 \tab 1 \tab 28 \tab 13 \tab 7 \tab 1 \tab 38 \tab 1 \tab - \tab 3 \tab 1 \tab 9 \tab 6 \tab 6 \tab - \tab - \tab - \tab 3 \tab - \tab 5 \tab - \tab 2 \tab - \tab 1 \tab - \tab 5 \tab 9 \tab 1 \tab - \tab - \tab 6 \tab 3 \tab 2 \tab 1 \tab 2 \tab 12 \tab 1 \tab 5 \tab 6 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab 5 \tab 3 \tab 1 \tab 1 \tab 3 \tab 4 \tab 3 \tab 1 \tab 12 \tab - \tab - \tab 10 \tab 12 \tab - \tab 1 \tab - \tab 2 \tab - \tab 5 \tab - \tab 7 \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 19 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 6 \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 6 \tab 3 \tab - \tab - \tab - \tab - \tab 4 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2867 \cr
   2018 \tab 6 \tab 6 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 7 \tab 7 \tab 3 \tab 13 \tab 4 \tab 3 \tab 7 \tab 10 \tab 1 \tab 9 \tab 6 \tab - \tab 3 \tab 4 \tab 5 \tab 1 \tab 1 \tab - \tab 13 \tab 3 \tab 5 \tab 20 \tab 2 \tab 2 \tab 65 \tab 3 \tab 6 \tab 3 \tab 2 \tab 4 \tab - \tab 2 \tab - \tab 10 \tab 2 \tab 2 \tab 2 \tab 81 \tab 8 \tab 6 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab 4 \tab 4 \tab 12 \tab - \tab 3 \tab 49 \tab - \tab 1 \tab 1378 \tab 5 \tab 9 \tab 3 \tab 18 \tab 1 \tab 2 \tab 22 \tab 1 \tab 1 \tab 14 \tab 1 \tab - \tab 4 \tab 3 \tab - \tab 7 \tab 4 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 13 \tab 15 \tab 3 \tab 2 \tab 17 \tab 9 \tab - \tab 1 \tab - \tab 4 \tab - \tab 3 \tab 15 \tab 1 \tab 5 \tab 1 \tab 13 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 5 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab 2 \tab 6 \tab 1 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab 1 \tab 6 \tab 6 \tab 1 \tab 15 \tab 10 \tab 2 \tab 1 \tab 30 \tab - \tab - \tab 4 \tab 2 \tab 7 \tab 8 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab 6 \tab 17 \tab 1 \tab - \tab 2 \tab 6 \tab 6 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 11 \tab 6 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 6 \tab - \tab - \tab 6 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 6 \tab - \tab - \tab 6 \tab 9 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 4 \tab - \tab 9 \tab 4 \tab 1 \tab - \tab 3 \tab 1 \tab 4 \tab 11 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab 9 \tab 8 \tab 3 \tab - \tab 4 \tab 9 \tab 10 \tab 6 \tab 6 \tab - \tab 23 \tab 3 \tab 14 \tab 19 \tab 32 \tab 1 \tab 9 \tab 4 \tab - \tab 11 \tab 3 \tab 4 \tab 3 \tab 2 \tab 6 \tab 17 \tab 14 \tab 7 \tab 58 \tab 2 \tab 3 \tab 123 \tab 1 \tab 16 \tab 4 \tab 2 \tab 5 \tab 4 \tab 4 \tab - \tab 7 \tab 1 \tab 5 \tab 7 \tab 163 \tab 2 \tab 28 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 7 \tab 8 \tab 1 \tab 3 \tab 6 \tab 24 \tab 2 \tab 8 \tab 90 \tab 3 \tab 1 \tab 2024 \tab 7 \tab 14 \tab 10 \tab 21 \tab 5 \tab 3 \tab 45 \tab 3 \tab - \tab 22 \tab 3 \tab 2 \tab 7 \tab 9 \tab 7 \tab 16 \tab 12 \tab - \tab - \tab 1 \tab 7 \tab 5 \tab 13 \tab 2 \tab - \tab 3 \tab 2 \tab 18 \tab 7 \tab 2 \tab 2 \tab 50 \tab 7 \tab 1 \tab 3 \tab 1 \tab 11 \tab - \tab 2 \tab 19 \tab - \tab 16 \tab 1 \tab 19 \tab 1 \tab 2 \tab 14 \tab 2 \tab - \tab 5 \tab 5 \tab 9 \tab 12 \tab 1 \tab 1 \tab 6 \tab 3 \tab 1 \tab 1 \tab 12 \tab 6 \tab 5 \tab 1 \tab 9 \tab 4 \tab 4 \tab 5 \tab 3 \tab 5 \tab 4 \tab 4 \tab - \tab 35 \tab 9 \tab 9 \tab 2 \tab 33 \tab 1 \tab 1 \tab 13 \tab 6 \tab 13 \tab 19 \tab 8 \tab 2 \tab - \tab - \tab 4 \tab 2 \tab 4 \tab 2 \tab 4 \tab 3 \tab - \tab 1 \tab 7 \tab 21 \tab - \tab - \tab - \tab 11 \tab 3 \tab 1 \tab 4 \tab 1 \tab 10 \tab 22 \tab 5 \tab 13 \tab 3 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 5 \tab 18 \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab 1 \tab 5 \tab 8 \tab - \tab 3 \tab 11 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 24 \tab - \tab 1 \tab 25 \tab 18 \tab 2 \tab 6 \tab - \tab 3 \tab 4 \tab 9 \tab - \tab 4 \tab - \tab - \tab 2 \tab 5 \tab - \tab 2 \tab 43 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 4 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 8 \tab 2 \tab 1 \tab 2 \tab 6 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab 3 \tab 10 \tab 3 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 125 \tab 4032 \cr
   2022 \tab 8 \tab 7 \tab 5 \tab - \tab - \tab 2 \tab - \tab 6 \tab 5 \tab - \tab 22 \tab 3 \tab 9 \tab 11 \tab 17 \tab 2 \tab 2 \tab 5 \tab - \tab 5 \tab 8 \tab 8 \tab 2 \tab 1 \tab 1 \tab 11 \tab 11 \tab 7 \tab 36 \tab 1 \tab 1 \tab 90 \tab 1 \tab 7 \tab - \tab 3 \tab 6 \tab 1 \tab 2 \tab 2 \tab 6 \tab - \tab 4 \tab 4 \tab 116 \tab - \tab 24 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 5 \tab 1 \tab 1 \tab 6 \tab 14 \tab - \tab 1 \tab 69 \tab - \tab - \tab 2080 \tab 5 \tab 12 \tab 4 \tab 39 \tab 3 \tab 4 \tab 30 \tab 1 \tab - \tab 7 \tab 1 \tab 2 \tab 9 \tab 1 \tab 2 \tab 12 \tab 5 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 6 \tab 1 \tab - \tab 3 \tab 1 \tab 18 \tab 9 \tab 2 \tab 2 \tab 24 \tab 8 \tab 2 \tab 4 \tab - \tab 13 \tab 1 \tab 3 \tab 18 \tab 1 \tab 5 \tab 2 \tab 18 \tab 2 \tab 2 \tab 11 \tab - \tab 2 \tab 1 \tab 5 \tab 6 \tab 8 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 10 \tab 6 \tab 12 \tab - \tab 4 \tab 6 \tab 2 \tab 5 \tab 3 \tab 2 \tab 3 \tab 5 \tab - \tab 22 \tab 13 \tab 6 \tab 4 \tab 29 \tab - \tab 1 \tab 7 \tab 2 \tab 9 \tab 6 \tab 16 \tab 3 \tab 1 \tab 4 \tab 4 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 10 \tab 17 \tab - \tab - \tab - \tab 8 \tab 2 \tab - \tab 5 \tab - \tab 10 \tab 19 \tab 3 \tab 7 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab 3 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 7 \tab - \tab 1 \tab 15 \tab 1 \tab 1 \tab 2 \tab 6 \tab - \tab 2 \tab 5 \tab 9 \tab - \tab 6 \tab 19 \tab 14 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 6 \tab - \tab 3 \tab - \tab - \tab 1 \tab 3 \tab - \tab 5 \tab 25 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 5 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 6 \tab 6 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 88 \tab 3544 \cr
   Total \tab 197 \tab 112 \tab 108 \tab 13 \tab 39 \tab 132 \tab 221 \tab 260 \tab 146 \tab 37 \tab 271 \tab 103 \tab 257 \tab 292 \tab 560 \tab 34 \tab 233 \tab 77 \tab 10 \tab 182 \tab 117 \tab 80 \tab 32 \tab 29 \tab 58 \tab 273 \tab 152 \tab 197 \tab 663 \tab 80 \tab 129 \tab 2433 \tab 65 \tab 307 \tab 302 \tab 61 \tab 506 \tab 53 \tab 120 \tab 38 \tab 270 \tab 40 \tab 178 \tab 162 \tab 2454 \tab 86 \tab 299 \tab 112 \tab 30 \tab 21 \tab 44 \tab 14 \tab 60 \tab 40 \tab 151 \tab 182 \tab 146 \tab 79 \tab 195 \tab 527 \tab 43 \tab 45 \tab 1509 \tab 37 \tab 31 \tab 36760 \tab 130 \tab 191 \tab 235 \tab 709 \tab 247 \tab 68 \tab 640 \tab 62 \tab 17 \tab 323 \tab 57 \tab 65 \tab 206 \tab 87 \tab 37 \tab 133 \tab 121 \tab 39 \tab 16 \tab 67 \tab 98 \tab 98 \tab 130 \tab 32 \tab 44 \tab 77 \tab 48 \tab 591 \tab 438 \tab 56 \tab 118 \tab 776 \tab 161 \tab 31 \tab 56 \tab 20 \tab 210 \tab 21 \tab 47 \tab 364 \tab 30 \tab 187 \tab 21 \tab 346 \tab 29 \tab 38 \tab 214 \tab 59 \tab 34 \tab 35 \tab 58 \tab 71 \tab 83 \tab 32 \tab 55 \tab 98 \tab 72 \tab 45 \tab 86 \tab 100 \tab 145 \tab 261 \tab 28 \tab 187 \tab 267 \tab 77 \tab 116 \tab 83 \tab 85 \tab 88 \tab 196 \tab 14 \tab 535 \tab 230 \tab 130 \tab 69 \tab 887 \tab 43 \tab 59 \tab 120 \tab 36 \tab 183 \tab 164 \tab 152 \tab 26 \tab 17 \tab 50 \tab 62 \tab 43 \tab 115 \tab 33 \tab 41 \tab 39 \tab 31 \tab 17 \tab 223 \tab 501 \tab 82 \tab 435 \tab 13 \tab 117 \tab 96 \tab 26 \tab 74 \tab 33 \tab 327 \tab 53 \tab 146 \tab 171 \tab 52 \tab 46 \tab 8 \tab 61 \tab 21 \tab 110 \tab 22 \tab 80 \tab 13 \tab 101 \tab 20 \tab 42 \tab 47 \tab 24 \tab 5 \tab 43 \tab 29 \tab 71 \tab 96 \tab 51 \tab 39 \tab 134 \tab 41 \tab 17 \tab 34 \tab 63 \tab 50 \tab 54 \tab 53 \tab 176 \tab 25 \tab 30 \tab 134 \tab 170 \tab 21 \tab 96 \tab 30 \tab 40 \tab 61 \tab 70 \tab 21 \tab 92 \tab 56 \tab 73 \tab 12 \tab 62 \tab 9 \tab 65 \tab 255 \tab 41 \tab 13 \tab 23 \tab 22 \tab 21 \tab 59 \tab 8 \tab 100 \tab 12 \tab 76 \tab 43 \tab 36 \tab 32 \tab 21 \tab 12 \tab 20 \tab 39 \tab 10 \tab 23 \tab 35 \tab 44 \tab 25 \tab 31 \tab 15 \tab 8 \tab 22 \tab 31 \tab 13 \tab 36 \tab 18 \tab 54 \tab 117 \tab 37 \tab 10 \tab 15 \tab 10 \tab 30 \tab 47 \tab 18 \tab 1 \tab 5 \tab 16 \tab 4 \tab 3 \tab 9 \tab 1 \tab 6 \tab 1 \tab 213 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{170} crop production
\item \code{180} animal production
\item \code{190} forestry, except logging
\item \code{270} logging
\item \code{280} fishing, hunting, and trapping
\item \code{290} support activities for agriculture and forestry
\item \code{370} oil and gas extraction
\item \code{380} coal mining
\item \code{390} metal ore mining
\item \code{470} nonmetallic mineral mining and quarrying
\item \code{480} not specified type of mining
\item \code{490} support activities for mining
\item \code{570} electric power generation, transmission, and distribution
\item \code{580} natural gas distribution
\item \code{590} electric and gas, and other combinations
\item \code{670} water, steam, air-conditioning, and irrigation systems
\item \code{680} sewage treatment facilities
\item \code{690} not specified utilities
\item \code{770} construction
\item \code{1070} animal food, grain, and oilseed milling
\item \code{1080} sugar and confectionery products
\item \code{1090} fruit and vegetable preserving and specialty food manufacturing
\item \code{1170} dairy product manufacturing
\item \code{1180} animal slaughtering and processing
\item \code{1190} retail bakeries
\item \code{1270} bakeries, except retail
\item \code{1280} seafood and other miscellaneous foods, n.e.c.
\item \code{1290} not specified food industries
\item \code{1370} beverage manufacturing
\item \code{1390} tobacco manufacturing
\item \code{1470} fiber, yarn, and thread mills
\item \code{1480} fabric mills, except knitting mills
\item \code{1490} textile and fabric finishing and coating mills
\item \code{1570} carpet and rug mills
\item \code{1590} textile product mills, except carpet and rug
\item \code{1670} knitting fabric mills, and apparel knitting mills
\item \code{1680} cut and sew apparel manufacturing
\item \code{1690} apparel accessories and other apparel manufacturing
\item \code{1770} footwear manufacturing
\item \code{1790} leather tanning and finishing and other allied products manufacturing
\item \code{1870} pulp, paper, and paperboard mills
\item \code{1880} paperboard containers and boxes
\item \code{1890} miscellaneous paper and pulp products
\item \code{1990} printing and related support activities
\item \code{2070} petroleum refining
\item \code{2090} miscellaneous petroleum and coal products
\item \code{2170} resin, synthetic rubber and fibers, and filaments manufacturing
\item \code{2180} agricultural chemical manufacturing
\item \code{2190} pharmaceutical and medicine manufacturing
\item \code{2270} paint, coating, and adhesive manufacturing
\item \code{2280} soap, cleaning compound, and cosmetics manufacturing
\item \code{2290} industrial and miscellaneous chemicals
\item \code{2370} plastics product manufacturing
\item \code{2380} tire manufacturing
\item \code{2390} rubber product, except tire, manufacturing
\item \code{2470} pottery, ceramics, and plumbing fixture manufacturing
\item \code{2480} structural clay product manufacturing
\item \code{2490} glass and glass product manufacturing
\item \code{2570} cement, concrete, lime, and gypsum product manufacturing
\item \code{2590} miscellaneous nonmetallic mineral product manufacturing
\item \code{2670} iron and steel mills and steel product manufacturing
\item \code{2680} aluminum production and processing
\item \code{2690} nonferrous metal (except aluminum) production and processing
\item \code{2770} foundries
\item \code{2780} metal forgings and stampings
\item \code{2790} cutlery and hand tool manufacturing
\item \code{2870} structural metals, and boiler, tank, and shipping container manufacturing
\item \code{2880} machine shops; turned product; screw, nut, and bolt manufacturing
\item \code{2890} coating, engraving, heat treating and allied activities
\item \code{2970} ordnance
\item \code{2980} miscellaneous fabricated metal products manufacturing
\item \code{2990} not specified metal industries
\item \code{3070} agricultural implement manufacturing
\item \code{3080} construction, and mining and oil and gas field machinery manufacturing
\item \code{3090} commercial and service industry machinery manufacturing
\item \code{3170} metalworking machinery manufacturing
\item \code{3180} engines, turbines, and power transmission equipment manufacturing
\item \code{3190} machinery manufacturing, n.e.c.
\item \code{3290} not specified machinery manufacturing
\item \code{3360} computer and peripheral equipment manufacturing
\item \code{3370} communications, and audio and video equipment manufacturing
\item \code{3380} navigational, measuring, electromedical, and control instruments manufacturing
\item \code{3390} electronic component and product manufacturing, n.e.c.
\item \code{3470} household appliance manufacturing
\item \code{3490} electrical lighting and electrical equipment manufacturing, and other electrical component manufacturing, n.e.c.
\item \code{3570} motor vehicles and motor vehicle equipment manufacturing
\item \code{3580} aircraft and parts manufacturing
\item \code{3590} aerospace product and parts manufacturing
\item \code{3670} railroad rolling stock manufacturing
\item \code{3680} ship and boat building
\item \code{3690} other transportation equipment manufacturing
\item \code{3770} sawmills and wood preservation
\item \code{3780} veneer, plywood, and engineered wood products
\item \code{3790} prefabricated wood buildings and mobile homes
\item \code{3870} miscellaneous wood products
\item \code{3890} furniture and related product manufacturing
\item \code{3960} medical equipment and supplies manufacturing
\item \code{3970} sporting and athletic goods, and doll, toy and game manufacturing
\item \code{3980} miscellaneous manufacturing, n.e.c.
\item \code{3990} not specified manufacturing industries
\item \code{4070} motor vehicles, parts and supplies, merchant wholesalers
\item \code{4080} furniture and home furnishing, merchant wholesalers
\item \code{4090} lumber and other construction materials, merchant wholesalers
\item \code{4170} professional and commercial equipment and supplies, merchant wholesalers
\item \code{4180} metals and minerals, except petroleum, merchant wholesalers
\item \code{4190} electrical and electronic goods, merchant wholesalers
\item \code{4260} hardware, plumbing and heating equipment, and supplies, merchant wholesalers
\item \code{4270} machinery, equipment, and supplies, merchant wholesalers
\item \code{4280} recyclable material, merchant wholesalers
\item \code{4290} miscellaneous durable goods, merchant wholesalers
\item \code{4370} paper and paper products, merchant wholesalers
\item \code{4380} drugs, sundries, and chemical and allied products, merchant wholesalers
\item \code{4390} apparel, fabrics, and notions, merchant wholesalers
\item \code{4470} groceries and related products, merchant wholesalers
\item \code{4480} farm product raw materials, merchant wholesalers
\item \code{4490} petroleum and petroleum products, merchant wholesalers
\item \code{4560} alcoholic beverages, merchant wholesalers
\item \code{4570} farm supplies, merchant wholesalers
\item \code{4580} miscellaneous nondurable goods, merchant wholesalers
\item \code{4585} wholesale electronic markets, agents and brokers
\item \code{4590} not specified wholesale trade
\item \code{4670} automobile dealers
\item \code{4680} other motor vehicle dealers
\item \code{4690} auto parts, accessories, and tire stores
\item \code{4770} furniture and home furnishings stores
\item \code{4780} household appliance stores
\item \code{4790} radio, tv, and computer stores
\item \code{4870} building material and supplies dealers
\item \code{4880} hardware stores
\item \code{4890} lawn and garden equipment and supplies stores
\item \code{4970} grocery stores
\item \code{4980} specialty food stores
\item \code{4990} beer, wine, and liquor stores
\item \code{5070} pharmacies and drug stores
\item \code{5080} health and personal care, except drug, stores
\item \code{5090} gasoline stations
\item \code{5170} clothing stores
\item \code{5180} shoe stores
\item \code{5190} jewelry, luggage, and leather goods stores
\item \code{5270} sporting goods, camera, and hobby and toy stores
\item \code{5280} sewing, needlework, and piece goods stores
\item \code{5290} music stores
\item \code{5370} book stores and news dealers
\item \code{5380} department stores and discount stores
\item \code{5390} miscellaneous general merchandise stores
\item \code{5470} retail florists
\item \code{5480} office supplies and stationery stores
\item \code{5490} used merchandise stores
\item \code{5570} gift, novelty, and souvenir shops
\item \code{5580} miscellaneous retail stores
\item \code{5590} electronic shopping
\item \code{5591} electronic auctions
\item \code{5592} mail order houses
\item \code{5670} vending machine operators
\item \code{5680} fuel dealers
\item \code{5690} other direct selling establishments
\item \code{5790} not specified retail trade
\item \code{6070} air transportation
\item \code{6080} rail transportation
\item \code{6090} water transportation
\item \code{6170} truck transportation
\item \code{6180} bus service and urban transit
\item \code{6190} taxi and limousine service
\item \code{6270} pipeline transportation
\item \code{6280} scenic and sightseeing transportation
\item \code{6290} services incidental to transportation
\item \code{6370} postal service
\item \code{6380} couriers and messengers
\item \code{6390} warehousing and storage
\item \code{6470} newspaper publishers
\item \code{6480} periodical, book, and directory publishers
\item \code{6490} software publishers
\item \code{6570} motion pictures and video industries
\item \code{6590} sound recording industries
\item \code{6670} radio and television broadcasting and cable subscription programming
\item \code{6672} internet publishing and broadcasting and web search portals
\item \code{6680} wired telecommunications carriers
\item \code{6690} other telecommunications services
\item \code{6695} data processing, hosting, and related services
\item \code{6770} libraries and archives
\item \code{6780} other information services
\item \code{6870} banking and related activities
\item \code{6880} savings institutions, including credit unions
\item \code{6890} non-depository credit and related activities
\item \code{6970} securities, commodities, funds, trusts, and other financial investments
\item \code{6990} insurance carriers and related activities
\item \code{7070} real estate
\item \code{7080} automotive equipment rental and leasing
\item \code{7170} video tape and disk rental
\item \code{7180} other consumer goods rental
\item \code{7190} commercial, industrial, and other intangible assets rental and leasing
\item \code{7270} legal services
\item \code{7280} accounting, tax preparation, bookkeeping, and payroll services
\item \code{7290} architectural, engineering, and related services
\item \code{7370} specialized design services
\item \code{7380} computer systems design and related services
\item \code{7390} management, scientific, and technical consulting services
\item \code{7460} scientific research and development services
\item \code{7470} advertising and related services
\item \code{7480} veterinary services
\item \code{7490} other professional, scientific, and technical services
\item \code{7570} management of companies and enterprises
\item \code{7580} employment services
\item \code{7590} business support services
\item \code{7670} travel arrangements and reservation services
\item \code{7680} investigation and security services
\item \code{7690} services to buildings and dwellings
\item \code{7770} landscaping services
\item \code{7780} other administrative and other support services
\item \code{7790} waste management and remediation services
\item \code{7860} elementary and secondary schools
\item \code{7870} colleges and universities, including junior colleges
\item \code{7880} business, technical, and trade schools and training
\item \code{7890} other schools and instruction, and educational support services
\item \code{7970} offices of physicians
\item \code{7980} offices of dentists
\item \code{7990} offices of chiropractors
\item \code{8070} offices of optometrists
\item \code{8080} offices of other health practitioners
\item \code{8090} outpatient care centers
\item \code{8170} home health care services
\item \code{8180} other health care services
\item \code{8190} hospitals
\item \code{8270} nursing care facilities
\item \code{8290} residential care facilities, without nursing
\item \code{8370} individual and family services
\item \code{8380} community food and housing, and emergency services
\item \code{8390} vocational rehabilitation services
\item \code{8470} child day care services
\item \code{8560} independent artists, performing arts, spectator sports, and related industries
\item \code{8570} museums, art galleries, historical sites, and similar institutions
\item \code{8580} bowling centers
\item \code{8590} other amusement, gambling, and recreation industries
\item \code{8660} traveler accommodation
\item \code{8670} recreational vehicle parks and camps, and rooming and boarding houses
\item \code{8680} restaurants and other food services
\item \code{8690} drinking places, alcoholic beverages
\item \code{8770} automotive repair and maintenance
\item \code{8780} car washes
\item \code{8790} electronic and precision equipment repair and maintenance
\item \code{8870} commercial and industrial machinery and equipment repair and maintenance
\item \code{8880} personal and household goods repair and maintenance
\item \code{8890} footwear and leather goods repair
\item \code{8970} barber shops
\item \code{8980} beauty salons
\item \code{8990} nail salons and other personal care services
\item \code{9070} drycleaning and laundry services
\item \code{9080} funeral homes, and cemeteries and crematories
\item \code{9090} other personal services
\item \code{9160} religious organizations
\item \code{9170} civic, social, advocacy organizations, and grantmaking and giving services
\item \code{9180} labor unions
\item \code{9190} business, professional, political, and similar organizations
\item \code{9290} private households
\item \code{9370} executive offices and legislative bodies
\item \code{9380} public finance activities
\item \code{9390} other general government and support
\item \code{9470} justice, public order, and safety activities
\item \code{9480} administration of human resource programs
\item \code{9490} administration of environmental quality and housing programs
\item \code{9570} administration of economic programs and space research
\item \code{9590} national security and international affairs
\item \code{9670} u.s. army
\item \code{9680} u.s. air force
\item \code{9690} u.s. navy
\item \code{9770} u.s. marine corps
\item \code{9780} u.s. coast guard
\item \code{9790} u.s. armed forces/n.s.
\item \code{9870} active duty military personnel, all other
\item \code{9890} uncodeable
\item \code{9997} uncodeable
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
