% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_05.R
\name{spind80}
\alias{spind80}
\title{spind80}
\source{
General Social Survey https://gss.norc.org
}
\description{
Spouse's industry code (1980)
}
\details{
Question 7. Spouse's occupation, prestige, and industry.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab accounting, auditing, and bookkeeping services \tab administration of environmental quality and household programs \tab administration of human resources programs \tab advertising \tab agricultural production, crops \tab agricultural production, livestock \tab agricultural services, except horticultural \tab air transportation \tab aircraft and parts \tab alcoholic beverages \tab apparel and accessories, except knit \tab apparel and accessory stores, except shoe \tab apparel, fabrics, and notions \tab auto and home supply stores \tab automotive repair shops \tab automotive services, except repair \tab banking \tab barber shops \tab beauty shops \tab beverage industries \tab blast furnaces, steelworks, rolling and finishing mills \tab book and stationery stores \tab bus service and urban transit \tab business management and consulting services \tab business services, n.e.c. \tab canned and preserved fruits and vegetables \tab cement, concrete, gypsum, and plaster products \tab child day care services \tab coal mining \tab colleges and universities \tab commercial research, development, and testing labs \tab computer and data processing services \tab construction \tab construction and material handling machines \tab credit agencies, n.e.c. \tab cutlery, hand tools, and other hardware \tab cycles and miscellaneous transportation equipment \tab dairy products \tab department stores \tab don't know \tab drug stores \tab eating and drinking places \tab educational services, n.e.c. \tab electric and gas, and other combinations \tab electric light and power \tab electrical goods \tab electrical machinery, equipment, and supplies, n.e.c. \tab electronic computing equipment \tab elementary and secondary schools \tab engineering, architectural, and surveying services \tab engines and turbines \tab fabricated structural metal products \tab farm products-raw materials \tab farm supplies \tab fishing, hunting, and trapping, etc. \tab food stores, n.e.c. \tab forestry \tab fuel and ice dealers \tab funeral service and crematories \tab furniture and fixtures \tab furniture and home furnishings \tab furniture and home furnishings stores \tab gas and steam supply systems \tab gasoline service stations \tab general government, n.e.c. \tab glass and glass products \tab grain mill products \tab groceries and related products \tab grocery stores \tab guided missiles, space vehicles, and parts \tab hardware stores \tab hardware, plumbing and heating supplies \tab health services, n.e.c. \tab horticultural services \tab hospitals \tab hotels and motels \tab household appliances \tab household appliances, tv, and radio stores \tab insurance \tab iron and steel foundries \tab justice, public order, and safety \tab knitting mills \tab laundry, cleaning, and garment services \tab legal services \tab liquor stores \tab lodging places, except hotels and motels \tab logging \tab lumber and building material retailing \tab lumber and construction materials \tab machinery, equipment, and supplies \tab machinery, except electrical, n.e.c. \tab meat products \tab membership organizations \tab metal forgings and stampings \tab metalworking machinery \tab miscellaneous entertainment and recreation services \tab miscellaneous fabricated metal products \tab miscellaneous general merchandise stores \tab miscellaneous manufacturing industries \tab miscellaneous paper and pulp products \tab miscellaneous personal services \tab miscellaneous plastics products \tab miscellaneous professional and related services \tab miscellaneous repair services \tab miscellaneous retail stores \tab miscellaneous textile mill products \tab miscellaneous vehicle dealers \tab miscellaneous wholesale, durable goods \tab miscellaneous wholesale, nondurable goods \tab miscellaneous wood products \tab mobile home dealers \tab motor vehicle dealers \tab motor vehicles and equipment \tab motor vehicles and motor vehicle equipment \tab national security and international affairs \tab newspaper publishing and printing \tab no answer \tab not specified electrical machinery, equipment, and supplies \tab not specified food industries \tab not specified manufacturing industries \tab not specified metal industries \tab not specified wholesale trade \tab nursing and personal care facilities \tab offices of chiropractors \tab offices of dentists \tab offices of optometrists \tab offices of physicians \tab other primary metal industries \tab paperboard containers and boxes \tab personnel supply services \tab petroleum and coal products \tab petroleum products \tab petroleum refining \tab photographic equipment and supplies \tab plastics, synthetics, and resins \tab pottery and related products \tab primary aluminum industries \tab printing, publishing, and allied industries, except newspapers \tab private households \tab public finance, taxation, and monetary policy \tab pulp, paper, and paperboard mills \tab radio and television broadcasting \tab railroads \tab real estate, including real estate-insurance-law offices \tab religious organizations \tab retail bakeries \tab retail florists \tab sanitary services \tab savings and loan associations \tab sawmills, planning mills, and millwork \tab scientific and controlling instruments \tab scrap and waste materials \tab screw machine products \tab security, commodity brokerage, and investment companies \tab services incidental to transportation \tab services to dwellings and other buildings \tab ship and boat building and repairing \tab social services, n.e.c. \tab sugar and confectionery products \tab telephone (wire and radio) \tab theaters and motion pictures \tab tobacco manufactures \tab trucking service \tab u.s. postal service \tab variety stores \tab warehousing and storage \tab wood buildings and mobile homes \tab yarn, thread, and fabric mills \tab administration of economic programs \tab agricultural chemicals \tab dairy products stores \tab drugs \tab drugs, chemicals, and allied products \tab electrical repair shops \tab executive and legislative offices \tab farm machinery and equipment \tab leather products, except footwear \tab metal mining \tab miscellaneous fabricated textile products \tab miscellaneous nonmetallic mineral and stone products \tab nonmetallic mining and quarrying, except fuel \tab not specified machinery \tab not specified retail trade \tab offices of health practitioners, n.e.c. \tab optical and health services supplies \tab other rubber products, and plastics footwear and belting \tab radio, tv, and communication equipment \tab railroad locomotives and equipment \tab residential care facilities, without nursing \tab sewing, needlework, and piece goods stores \tab shoe stores \tab soaps and cosmetics \tab sporting goods, bicycles, and hobby stores \tab taxicab service \tab tires and inner tubes \tab toys, amusement, and sporting goods \tab vending machine operators \tab water supply and irrigation \tab bakery products \tab crude petroleum and natural gas extraction \tab detective and protective services \tab direct selling establishments \tab footwear, except rubber and plastic \tab job training and vocational rehabilitation services \tab not specified electrical and hardware products \tab sporting goods, toys, and hobby goods \tab structural clay products \tab business, trade, and vocational schools \tab dyeing and finishing textiles, except wool and knit goods \tab jewelry stores \tab mail order houses \tab metals and minerals, except petroleum \tab miscellaneous food preparations and kindred products \tab miscellaneous petroleum and coal products \tab noncommercial educational and scientific research \tab ordnance \tab telegraph and miscellaneous communication service \tab dressmaking shops \tab libraries \tab museums, art galleries and zoos \tab not specified utilities \tab office and accounting machines \tab paper and paper products \tab water transportation \tab bowling alleys, billiard and pool parlors \tab industrial and miscellaneous chemicals \tab pipe lines, except natural gas \tab floor coverings, except hard surface \tab watches, clocks, and clockwork operated devices \tab retail nurseries and garden stores \tab 942 \tab 951 \tab 99 \tab shoe repair shops \tab 19 \tab 259 \tab 379 \tab 558 \tab 804 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 742 \tab 2 \tab 1 \tab 1 \tab 1 \tab 4 \tab 6 \tab 2 \tab 4 \tab 6 \tab 1 \tab 7 \tab 4 \tab 2 \tab 2 \tab 4 \tab 2 \tab 11 \tab 2 \tab 8 \tab 2 \tab 7 \tab 2 \tab 3 \tab 1 \tab 6 \tab 3 \tab 1 \tab 5 \tab 3 \tab 14 \tab 4 \tab 4 \tab 42 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 18 \tab 3 \tab 3 \tab 24 \tab 2 \tab 4 \tab 7 \tab 3 \tab 5 \tab 4 \tab 53 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab 21 \tab 2 \tab 1 \tab 5 \tab 6 \tab 7 \tab 3 \tab 2 \tab 9 \tab 2 \tab 28 \tab 1 \tab 1 \tab 2 \tab 22 \tab 4 \tab 6 \tab 1 \tab 5 \tab 7 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 3 \tab 5 \tab 1 \tab 4 \tab 1 \tab 2 \tab 6 \tab 5 \tab 2 \tab 3 \tab 2 \tab 7 \tab 2 \tab 1 \tab 3 \tab 8 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 6 \tab 2 \tab 14 \tab 4 \tab 2 \tab 2 \tab 2 \tab 2 \tab 9 \tab 1 \tab 2 \tab 7 \tab 1 \tab 3 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 5 \tab 11 \tab 1 \tab 1 \tab 4 \tab 4 \tab 19 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 5 \tab 3 \tab 4 \tab 2 \tab 8 \tab 1 \tab 1 \tab 17 \tab 7 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 734 \tab 2 \tab 1 \tab 2 \tab 3 \tab 11 \tab 11 \tab 1 \tab 3 \tab 2 \tab 1 \tab 10 \tab 9 \tab 2 \tab 1 \tab 4 \tab - \tab 18 \tab 1 \tab 2 \tab 2 \tab 11 \tab 1 \tab 4 \tab 4 \tab 5 \tab 3 \tab 3 \tab 1 \tab 2 \tab 18 \tab 1 \tab 2 \tab 54 \tab 5 \tab - \tab 1 \tab - \tab 1 \tab 9 \tab 1 \tab 3 \tab 24 \tab 1 \tab - \tab 4 \tab 1 \tab 6 \tab 3 \tab 39 \tab 4 \tab 3 \tab 6 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 4 \tab 4 \tab 11 \tab 4 \tab - \tab 9 \tab 12 \tab 2 \tab 2 \tab 4 \tab 7 \tab 4 \tab 37 \tab 1 \tab - \tab 2 \tab 19 \tab 2 \tab 9 \tab 1 \tab 2 \tab 10 \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab 3 \tab 11 \tab 3 \tab 2 \tab 4 \tab - \tab 5 \tab 4 \tab 5 \tab 3 \tab 2 \tab 1 \tab 3 \tab 3 \tab 2 \tab 8 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 6 \tab 2 \tab 11 \tab 29 \tab 8 \tab 13 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 12 \tab 2 \tab 4 \tab - \tab 7 \tab 3 \tab 1 \tab 2 \tab 6 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 10 \tab 8 \tab 5 \tab 2 \tab 2 \tab 7 \tab 12 \tab 8 \tab 1 \tab - \tab 4 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 5 \tab 1 \tab - \tab 1 \tab 10 \tab 3 \tab - \tab 11 \tab 5 \tab 1 \tab 2 \tab - \tab - \tab 4 \tab 3 \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 684 \tab 3 \tab 3 \tab 3 \tab 1 \tab 6 \tab 9 \tab 4 \tab 6 \tab 6 \tab 1 \tab 12 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 10 \tab - \tab 8 \tab 1 \tab 2 \tab 1 \tab 3 \tab 3 \tab 8 \tab 1 \tab - \tab 4 \tab 3 \tab 18 \tab 1 \tab 3 \tab 61 \tab 4 \tab 2 \tab 2 \tab - \tab - \tab 10 \tab - \tab 4 \tab 19 \tab 3 \tab - \tab 5 \tab 1 \tab 5 \tab 3 \tab 44 \tab 5 \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 4 \tab 1 \tab 3 \tab - \tab 2 \tab 12 \tab 3 \tab - \tab 5 \tab 14 \tab 2 \tab 2 \tab - \tab 6 \tab 1 \tab 22 \tab 4 \tab - \tab - \tab 19 \tab 3 \tab 7 \tab 2 \tab 2 \tab 10 \tab - \tab 5 \tab 2 \tab 1 \tab 1 \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 4 \tab 2 \tab 6 \tab 1 \tab - \tab 2 \tab 2 \tab 4 \tab 6 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 5 \tab 14 \tab 16 \tab 2 \tab 14 \tab 4 \tab - \tab - \tab - \tab - \tab 6 \tab 1 \tab 3 \tab - \tab 6 \tab - \tab 2 \tab 2 \tab 6 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 6 \tab 7 \tab 6 \tab 2 \tab 1 \tab 3 \tab 14 \tab 9 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 6 \tab - \tab 6 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 5 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 3 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 748 \tab 5 \tab 7 \tab 3 \tab 2 \tab 3 \tab 4 \tab 3 \tab 2 \tab 9 \tab - \tab 4 \tab 4 \tab - \tab 3 \tab 13 \tab 4 \tab 11 \tab - \tab 5 \tab 1 \tab 8 \tab 2 \tab 3 \tab 3 \tab 7 \tab 1 \tab 2 \tab 4 \tab 1 \tab 13 \tab 1 \tab 1 \tab 51 \tab 5 \tab 6 \tab 1 \tab - \tab 3 \tab 15 \tab 3 \tab 6 \tab 23 \tab 2 \tab 2 \tab 7 \tab 1 \tab 5 \tab 5 \tab 51 \tab 6 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 4 \tab 8 \tab - \tab 1 \tab 5 \tab 20 \tab - \tab 3 \tab 3 \tab 2 \tab 2 \tab 32 \tab 7 \tab - \tab 1 \tab 17 \tab 3 \tab 15 \tab 2 \tab 3 \tab 5 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 7 \tab 3 \tab 4 \tab - \tab 4 \tab 5 \tab 3 \tab 3 \tab 2 \tab 1 \tab 1 \tab 5 \tab - \tab 3 \tab 9 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 4 \tab 3 \tab 18 \tab 14 \tab 3 \tab 8 \tab 1 \tab - \tab 3 \tab - \tab - \tab 13 \tab - \tab 2 \tab 1 \tab 3 \tab 2 \tab - \tab 5 \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 7 \tab 7 \tab 4 \tab - \tab 3 \tab 2 \tab 11 \tab 6 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 4 \tab 4 \tab 3 \tab - \tab 8 \tab 6 \tab - \tab 16 \tab 10 \tab - \tab 2 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 781 \tab 2 \tab 2 \tab 4 \tab 5 \tab 9 \tab 6 \tab 1 \tab 6 \tab 4 \tab - \tab 16 \tab 6 \tab - \tab - \tab 2 \tab 1 \tab 19 \tab 1 \tab 4 \tab 5 \tab 2 \tab - \tab 6 \tab 1 \tab 7 \tab 1 \tab 1 \tab 5 \tab - \tab 17 \tab 3 \tab 3 \tab 63 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 12 \tab - \tab 2 \tab 24 \tab - \tab 1 \tab 4 \tab 1 \tab 8 \tab 5 \tab 40 \tab 5 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 4 \tab 3 \tab 4 \tab 13 \tab 2 \tab - \tab 8 \tab 15 \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab 42 \tab 5 \tab 1 \tab 3 \tab 17 \tab 3 \tab 18 \tab 1 \tab 2 \tab 15 \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 2 \tab 5 \tab 2 \tab 2 \tab - \tab 2 \tab 6 \tab 6 \tab 1 \tab 3 \tab 2 \tab 3 \tab 4 \tab 3 \tab 3 \tab 4 \tab 2 \tab - \tab 2 \tab 3 \tab 2 \tab - \tab 3 \tab 1 \tab 14 \tab 20 \tab 2 \tab 9 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 8 \tab 1 \tab 1 \tab - \tab 9 \tab 1 \tab 2 \tab 6 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 14 \tab 7 \tab 4 \tab 1 \tab 3 \tab 4 \tab 11 \tab 9 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 4 \tab 2 \tab 7 \tab 1 \tab 5 \tab - \tab 8 \tab 1 \tab - \tab 18 \tab 7 \tab 2 \tab 2 \tab - \tab 4 \tab 5 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1521 \tab 6 \tab 5 \tab 8 \tab 1 \tab 4 \tab 15 \tab 4 \tab 5 \tab 16 \tab - \tab 17 \tab 9 \tab 2 \tab 4 \tab 9 \tab 1 \tab 27 \tab 1 \tab 6 \tab 6 \tab 2 \tab 2 \tab 7 \tab 2 \tab 10 \tab 2 \tab 3 \tab 12 \tab - \tab 31 \tab - \tab 16 \tab 103 \tab 7 \tab 11 \tab 2 \tab 2 \tab 7 \tab 24 \tab - \tab 5 \tab 44 \tab 3 \tab 2 \tab 13 \tab 5 \tab 9 \tab 10 \tab 97 \tab 17 \tab 1 \tab 5 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 7 \tab - \tab 5 \tab 4 \tab 6 \tab 10 \tab 1 \tab 2 \tab 6 \tab 27 \tab 3 \tab 4 \tab 3 \tab 12 \tab 8 \tab 56 \tab 15 \tab 1 \tab 3 \tab 36 \tab 2 \tab 30 \tab 1 \tab 9 \tab 11 \tab 3 \tab 2 \tab 2 \tab 8 \tab 2 \tab 2 \tab 13 \tab 3 \tab 8 \tab - \tab 1 \tab 18 \tab 2 \tab 1 \tab 6 \tab 5 \tab - \tab 7 \tab 2 \tab 4 \tab 6 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab 7 \tab 2 \tab 28 \tab 42 \tab 7 \tab 13 \tab 4 \tab 1 \tab 4 \tab - \tab 3 \tab 25 \tab 2 \tab 10 \tab 2 \tab 16 \tab 2 \tab 2 \tab 11 \tab 13 \tab - \tab 7 \tab 6 \tab 1 \tab - \tab 1 \tab 12 \tab 9 \tab 4 \tab 2 \tab 7 \tab 7 \tab 28 \tab 5 \tab - \tab 1 \tab 7 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 3 \tab 11 \tab 1 \tab 10 \tab 2 \tab 10 \tab 7 \tab 2 \tab 22 \tab 10 \tab 1 \tab 2 \tab - \tab 4 \tab 4 \tab - \tab - \tab 7 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 5 \tab - \tab 7 \tab 4 \tab 3 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 6 \tab 2 \tab 5 \tab 4 \tab 3 \tab 4 \tab - \tab 5 \tab 2 \tab 2 \tab 1 \tab 5 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1563 \tab 12 \tab 4 \tab 7 \tab 2 \tab 11 \tab 9 \tab 4 \tab 7 \tab 7 \tab - \tab 12 \tab 9 \tab 1 \tab 5 \tab 8 \tab 1 \tab 22 \tab 2 \tab 7 \tab 3 \tab 2 \tab 2 \tab 2 \tab 8 \tab 7 \tab 2 \tab 5 \tab 10 \tab - \tab 18 \tab 3 \tab 16 \tab 81 \tab 1 \tab 4 \tab - \tab - \tab 4 \tab 15 \tab - \tab 5 \tab 46 \tab 3 \tab 2 \tab 8 \tab 3 \tab 11 \tab 14 \tab 95 \tab 10 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 7 \tab 1 \tab 6 \tab 2 \tab 1 \tab 6 \tab 1 \tab 4 \tab 8 \tab 17 \tab 3 \tab 1 \tab 3 \tab 9 \tab 2 \tab 61 \tab 9 \tab 3 \tab 2 \tab 21 \tab 3 \tab 21 \tab 2 \tab 2 \tab 16 \tab 1 \tab 5 \tab 1 \tab 8 \tab - \tab 6 \tab 15 \tab 1 \tab - \tab - \tab 6 \tab 12 \tab 4 \tab 7 \tab 5 \tab 3 \tab 1 \tab 8 \tab 2 \tab 3 \tab 10 \tab 5 \tab - \tab - \tab 3 \tab 1 \tab - \tab 10 \tab 1 \tab 22 \tab 37 \tab 5 \tab 17 \tab 6 \tab 2 \tab 11 \tab 1 \tab 4 \tab 18 \tab 4 \tab 6 \tab 2 \tab 14 \tab 4 \tab 2 \tab 8 \tab 7 \tab 3 \tab 12 \tab 3 \tab 2 \tab - \tab 1 \tab 13 \tab 6 \tab 2 \tab 6 \tab 3 \tab 11 \tab 21 \tab 18 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 10 \tab 3 \tab 7 \tab - \tab 9 \tab - \tab 13 \tab 10 \tab - \tab 26 \tab 7 \tab 4 \tab 4 \tab 1 \tab 4 \tab 9 \tab 2 \tab - \tab 3 \tab 1 \tab 5 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 5 \tab 3 \tab - \tab 4 \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab 4 \tab 4 \tab 4 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 7 \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1527 \tab 4 \tab 3 \tab 3 \tab 4 \tab 7 \tab 6 \tab 5 \tab 8 \tab 14 \tab 1 \tab 5 \tab 9 \tab 1 \tab 2 \tab 14 \tab 4 \tab 22 \tab - \tab 14 \tab 2 \tab 9 \tab 1 \tab 6 \tab 11 \tab 15 \tab - \tab 2 \tab 8 \tab 1 \tab 31 \tab 3 \tab 19 \tab 72 \tab 3 \tab 6 \tab 2 \tab 1 \tab 2 \tab 12 \tab - \tab 5 \tab 36 \tab 2 \tab 4 \tab 8 \tab - \tab 13 \tab 6 \tab 82 \tab 11 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 9 \tab 1 \tab 10 \tab 2 \tab 4 \tab 18 \tab 1 \tab - \tab 5 \tab 14 \tab 5 \tab 2 \tab 1 \tab 11 \tab 2 \tab 58 \tab 7 \tab 2 \tab 4 \tab 25 \tab 1 \tab 27 \tab 1 \tab 3 \tab 11 \tab - \tab 3 \tab - \tab 6 \tab 1 \tab 6 \tab 11 \tab 3 \tab 4 \tab 1 \tab 3 \tab 10 \tab 1 \tab 4 \tab 7 \tab 4 \tab 4 \tab 5 \tab 2 \tab 4 \tab 10 \tab 2 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 11 \tab 1 \tab 24 \tab 31 \tab 3 \tab 43 \tab 1 \tab 2 \tab 6 \tab - \tab - \tab 16 \tab - \tab 6 \tab 2 \tab 15 \tab 3 \tab 1 \tab 5 \tab 9 \tab 2 \tab 4 \tab - \tab 2 \tab - \tab 2 \tab 8 \tab 11 \tab 4 \tab 6 \tab 5 \tab 7 \tab 29 \tab 9 \tab - \tab - \tab 5 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab 12 \tab 6 \tab 5 \tab 1 \tab 11 \tab 2 \tab 14 \tab 5 \tab - \tab 25 \tab 10 \tab 2 \tab 5 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 6 \tab 2 \tab 3 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 2 \tab 5 \tab 1 \tab 6 \tab - \tab 3 \tab - \tab 2 \tab 3 \tab 4 \tab 3 \tab - \tab 3 \tab - \tab 3 \tab 5 \tab 4 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1608 \tab 5 \tab 2 \tab 1 \tab 8 \tab 8 \tab 12 \tab 5 \tab 7 \tab 11 \tab - \tab 11 \tab 7 \tab 1 \tab 4 \tab 11 \tab 3 \tab 23 \tab - \tab 6 \tab 4 \tab 6 \tab 3 \tab 3 \tab 7 \tab 10 \tab - \tab 2 \tab 6 \tab 1 \tab 24 \tab - \tab 17 \tab 76 \tab 1 \tab 11 \tab - \tab 1 \tab 4 \tab 17 \tab - \tab 7 \tab 28 \tab - \tab - \tab 6 \tab 1 \tab 6 \tab 5 \tab 90 \tab 12 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab 1 \tab 7 \tab - \tab 4 \tab 1 \tab 2 \tab 28 \tab 2 \tab 4 \tab 6 \tab 13 \tab 5 \tab 2 \tab - \tab 22 \tab 2 \tab 38 \tab 9 \tab 3 \tab 7 \tab 21 \tab - \tab 12 \tab - \tab 1 \tab 19 \tab 2 \tab 3 \tab 1 \tab 5 \tab 4 \tab 5 \tab 7 \tab 8 \tab - \tab - \tab 2 \tab 9 \tab 1 \tab - \tab 5 \tab 2 \tab 6 \tab 5 \tab 4 \tab 3 \tab 11 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 8 \tab - \tab 24 \tab 14 \tab 4 \tab 50 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 13 \tab 2 \tab 4 \tab 1 \tab 10 \tab 5 \tab - \tab 2 \tab 11 \tab - \tab 11 \tab 3 \tab 4 \tab - \tab 4 \tab 11 \tab 10 \tab 3 \tab 9 \tab 2 \tab 6 \tab 15 \tab 4 \tab - \tab 2 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab - \tab 6 \tab 2 \tab 9 \tab 3 \tab 15 \tab 3 \tab 10 \tab 6 \tab - \tab 27 \tab 6 \tab - \tab 9 \tab 1 \tab 4 \tab 3 \tab 3 \tab - \tab 4 \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 3 \tab - \tab 4 \tab - \tab 3 \tab 2 \tab 4 \tab - \tab 5 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 4 \tab 6 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 7 \tab 3 \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1547 \tab 13 \tab 3 \tab 4 \tab 1 \tab 4 \tab 9 \tab 3 \tab 13 \tab 7 \tab - \tab 7 \tab 5 \tab - \tab 2 \tab 13 \tab 4 \tab 23 \tab 1 \tab 10 \tab 3 \tab 7 \tab - \tab 2 \tab 4 \tab 14 \tab 1 \tab 1 \tab 13 \tab 1 \tab 22 \tab - \tab 21 \tab 76 \tab - \tab 4 \tab 3 \tab - \tab 3 \tab 14 \tab 1 \tab 5 \tab 40 \tab 3 \tab 1 \tab 9 \tab 3 \tab 6 \tab 7 \tab 94 \tab 10 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 7 \tab 2 \tab 1 \tab 13 \tab 4 \tab 5 \tab 9 \tab 15 \tab - \tab 1 \tab - \tab 28 \tab 7 \tab 54 \tab 13 \tab 1 \tab 5 \tab 18 \tab - \tab 24 \tab - \tab 5 \tab 9 \tab 1 \tab 6 \tab 1 \tab 3 \tab 1 \tab 8 \tab 16 \tab 2 \tab 6 \tab - \tab 4 \tab 8 \tab 1 \tab 1 \tab 5 \tab 2 \tab - \tab 4 \tab 2 \tab 2 \tab 5 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 6 \tab 1 \tab 17 \tab 19 \tab 9 \tab 26 \tab 2 \tab 2 \tab 7 \tab 1 \tab 3 \tab 11 \tab 2 \tab 7 \tab 1 \tab 23 \tab 4 \tab 2 \tab 3 \tab 11 \tab 1 \tab 7 \tab 2 \tab 3 \tab - \tab 2 \tab 9 \tab 4 \tab 8 \tab 3 \tab 2 \tab 2 \tab 22 \tab 7 \tab - \tab 3 \tab 10 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 7 \tab 5 \tab 8 \tab 1 \tab 10 \tab - \tab 12 \tab 4 \tab 1 \tab 23 \tab 7 \tab - \tab 4 \tab 1 \tab 4 \tab 4 \tab 1 \tab - \tab 4 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 2 \tab 8 \tab 1 \tab 4 \tab 1 \tab 5 \tab - \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 7 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1406 \tab 17 \tab 5 \tab 9 \tab 3 \tab 3 \tab 5 \tab 5 \tab 6 \tab 6 \tab 1 \tab 3 \tab 9 \tab - \tab 7 \tab 17 \tab 3 \tab 23 \tab 1 \tab 12 \tab - \tab 2 \tab 3 \tab 5 \tab 7 \tab 11 \tab 2 \tab 3 \tab 13 \tab 2 \tab 26 \tab 11 \tab 17 \tab 102 \tab 2 \tab 14 \tab 1 \tab - \tab - \tab 15 \tab - \tab 7 \tab 33 \tab 9 \tab - \tab 10 \tab 1 \tab 10 \tab 5 \tab 100 \tab 4 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 2 \tab 8 \tab 2 \tab 2 \tab 19 \tab 2 \tab 3 \tab 11 \tab 24 \tab - \tab 2 \tab - \tab 16 \tab 10 \tab 61 \tab 12 \tab 2 \tab 6 \tab 20 \tab 1 \tab 25 \tab 1 \tab 6 \tab 19 \tab - \tab 7 \tab 2 \tab 6 \tab - \tab 7 \tab 3 \tab 7 \tab 5 \tab - \tab 2 \tab 8 \tab 4 \tab - \tab 4 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 17 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 9 \tab 2 \tab 12 \tab 26 \tab 2 \tab 75 \tab 2 \tab 3 \tab 13 \tab 1 \tab 1 \tab 22 \tab 2 \tab 7 \tab - \tab 21 \tab 1 \tab - \tab 4 \tab 6 \tab 5 \tab 5 \tab 1 \tab 1 \tab - \tab 2 \tab 14 \tab 10 \tab 8 \tab 4 \tab 1 \tab 6 \tab 21 \tab 8 \tab - \tab 2 \tab 6 \tab - \tab 7 \tab - \tab 1 \tab 1 \tab 10 \tab 5 \tab 8 \tab 4 \tab 10 \tab 1 \tab 17 \tab 1 \tab 1 \tab 26 \tab 9 \tab 5 \tab 3 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab 6 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 5 \tab - \tab 4 \tab 2 \tab 5 \tab - \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 5 \tab - \tab 4 \tab 8 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 6 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2465 \tab 8 \tab 8 \tab 13 \tab 6 \tab 12 \tab 11 \tab 8 \tab 11 \tab 9 \tab 2 \tab 11 \tab 5 \tab 1 \tab 3 \tab 17 \tab 5 \tab 34 \tab - \tab 11 \tab 6 \tab 7 \tab 3 \tab 8 \tab 12 \tab 15 \tab 3 \tab 2 \tab 14 \tab - \tab 41 \tab 3 \tab 31 \tab 150 \tab 5 \tab 10 \tab - \tab 1 \tab 3 \tab 21 \tab - \tab 3 \tab 63 \tab 5 \tab - \tab 8 \tab 5 \tab 16 \tab 8 \tab 156 \tab 12 \tab 1 \tab 7 \tab - \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 2 \tab 13 \tab 1 \tab 14 \tab 3 \tab 4 \tab 31 \tab 6 \tab 1 \tab 8 \tab 36 \tab 8 \tab 2 \tab 4 \tab 23 \tab 13 \tab 90 \tab 12 \tab 3 \tab 2 \tab 35 \tab 1 \tab 43 \tab 3 \tab 6 \tab 10 \tab 1 \tab 3 \tab 4 \tab 7 \tab 3 \tab 4 \tab 10 \tab 6 \tab 4 \tab - \tab 4 \tab 11 \tab 3 \tab 2 \tab 4 \tab 2 \tab 7 \tab 7 \tab 11 \tab 6 \tab 18 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 18 \tab 1 \tab 29 \tab 38 \tab 7 \tab 74 \tab 3 \tab 3 \tab 10 \tab 2 \tab 1 \tab 21 \tab 2 \tab 14 \tab - \tab 24 \tab 1 \tab 1 \tab 9 \tab 16 \tab 6 \tab 4 \tab 2 \tab 6 \tab - \tab 2 \tab 15 \tab 29 \tab 4 \tab 5 \tab 4 \tab 6 \tab 31 \tab 20 \tab 2 \tab 4 \tab 7 \tab - \tab 6 \tab 2 \tab 3 \tab 1 \tab 19 \tab 12 \tab 10 \tab 3 \tab 11 \tab 1 \tab 17 \tab 10 \tab 1 \tab 54 \tab 10 \tab 1 \tab 3 \tab - \tab 2 \tab 20 \tab 1 \tab - \tab 9 \tab 3 \tab 4 \tab 4 \tab - \tab 1 \tab - \tab 4 \tab 2 \tab 2 \tab 1 \tab 5 \tab 3 \tab 8 \tab 4 \tab 5 \tab 1 \tab 6 \tab 2 \tab - \tab 2 \tab 4 \tab 1 \tab 2 \tab 2 \tab - \tab 4 \tab 2 \tab 6 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 5 \tab - \tab 2 \tab 1 \tab 10 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1080 \tab 5 \tab 1 \tab 5 \tab 4 \tab 4 \tab 9 \tab 4 \tab 6 \tab 6 \tab - \tab 4 \tab 4 \tab - \tab 4 \tab 5 \tab 3 \tab 19 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab 3 \tab 3 \tab 6 \tab 1 \tab 1 \tab 4 \tab 1 \tab 21 \tab 1 \tab 11 \tab 66 \tab - \tab 8 \tab 1 \tab - \tab 1 \tab 13 \tab - \tab 2 \tab 20 \tab 1 \tab - \tab 10 \tab 2 \tab 8 \tab 3 \tab 67 \tab 10 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab - \tab 4 \tab 1 \tab 2 \tab 22 \tab 3 \tab 2 \tab 3 \tab 7 \tab - \tab 2 \tab 6 \tab 9 \tab 7 \tab 41 \tab 15 \tab 1 \tab 1 \tab 15 \tab 2 \tab 24 \tab - \tab - \tab 7 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 5 \tab 4 \tab 2 \tab 5 \tab - \tab 3 \tab 11 \tab - \tab 2 \tab 4 \tab 3 \tab 1 \tab 3 \tab 4 \tab 2 \tab 7 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 10 \tab - \tab 6 \tab 17 \tab 2 \tab 33 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 16 \tab 1 \tab 11 \tab 1 \tab 9 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 7 \tab 17 \tab - \tab 1 \tab 2 \tab 4 \tab 12 \tab 12 \tab 1 \tab - \tab 5 \tab - \tab 5 \tab 1 \tab - \tab 1 \tab 7 \tab 6 \tab 4 \tab 3 \tab 7 \tab - \tab 11 \tab 1 \tab 1 \tab 14 \tab 6 \tab - \tab 1 \tab - \tab 3 \tab 5 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 4 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab - \tab 3 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1187 \tab 1 \tab 3 \tab 4 \tab 7 \tab 4 \tab 1 \tab 3 \tab 7 \tab 1 \tab 2 \tab 5 \tab 2 \tab - \tab 2 \tab 4 \tab 1 \tab 21 \tab - \tab 4 \tab 2 \tab 3 \tab 1 \tab 5 \tab 5 \tab 11 \tab 2 \tab - \tab 5 \tab - \tab 25 \tab 4 \tab 10 \tab 49 \tab - \tab 4 \tab 2 \tab - \tab 4 \tab 12 \tab - \tab 4 \tab 16 \tab 6 \tab - \tab 5 \tab 1 \tab 1 \tab 6 \tab 58 \tab 7 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab - \tab 3 \tab 1 \tab 5 \tab 10 \tab - \tab 1 \tab 2 \tab 11 \tab - \tab 2 \tab 4 \tab 10 \tab 5 \tab 41 \tab 5 \tab - \tab 2 \tab 14 \tab 1 \tab 13 \tab - \tab - \tab 10 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 4 \tab 4 \tab 3 \tab 6 \tab 1 \tab 1 \tab 8 \tab 2 \tab - \tab 9 \tab 1 \tab 3 \tab 3 \tab 1 \tab 3 \tab 7 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 6 \tab 1 \tab 9 \tab 20 \tab 2 \tab 33 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 11 \tab 1 \tab 7 \tab 3 \tab 11 \tab - \tab 2 \tab 7 \tab 2 \tab 1 \tab 6 \tab - \tab 3 \tab 2 \tab - \tab 5 \tab 11 \tab 2 \tab 2 \tab 3 \tab 1 \tab 15 \tab 6 \tab 1 \tab - \tab 2 \tab - \tab 5 \tab - \tab - \tab - \tab 6 \tab 1 \tab 3 \tab - \tab 12 \tab - \tab 7 \tab 3 \tab 1 \tab 20 \tab 8 \tab - \tab 4 \tab - \tab - \tab 7 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 4 \tab 2 \tab 4 \tab - \tab 2 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 10 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 56771 \tab 85 \tab 48 \tab 67 \tab 48 \tab 90 \tab 113 \tab 52 \tab 91 \tab 104 \tab 9 \tab 124 \tab 84 \tab 10 \tab 41 \tab 121 \tab 33 \tab 283 \tab 9 \tab 100 \tab 38 \tab 71 \tab 21 \tab 60 \tab 71 \tab 132 \tab 22 \tab 26 \tab 104 \tab 15 \tab 319 \tab 35 \tab 171 \tab 1046 \tab 37 \tab 85 \tab 18 \tab 7 \tab 33 \tab 207 \tab 8 \tab 61 \tab 440 \tab 40 \tab 16 \tab 104 \tab 28 \tab 109 \tab 84 \tab 1066 \tab 115 \tab 12 \tab 34 \tab 13 \tab 14 \tab 15 \tab 19 \tab 13 \tab 8 \tab 9 \tab 72 \tab 12 \tab 73 \tab 27 \tab 46 \tab 222 \tab 31 \tab 24 \tab 90 \tab 231 \tab 35 \tab 30 \tab 31 \tab 166 \tab 68 \tab 661 \tab 115 \tab 18 \tab 40 \tab 299 \tab 26 \tab 274 \tab 15 \tab 46 \tab 159 \tab 12 \tab 43 \tab 19 \tab 57 \tab 20 \tab 61 \tab 113 \tab 46 \tab 51 \tab 8 \tab 36 \tab 121 \tab 40 \tab 30 \tab 66 \tab 31 \tab 35 \tab 61 \tab 39 \tab 44 \tab 126 \tab 25 \tab 10 \tab 10 \tab 11 \tab 15 \tab 1 \tab 109 \tab 22 \tab 242 \tab 327 \tab 58 \tab 410 \tab 30 \tab 20 \tab 71 \tab 8 \tab 19 \tab 199 \tab 21 \tab 85 \tab 15 \tab 172 \tab 29 \tab 19 \tab 66 \tab 99 \tab 20 \tab 74 \tab 22 \tab 25 \tab 5 \tab 17 \tab 136 \tab 147 \tab 55 \tab 44 \tab 42 \tab 70 \tab 261 \tab 124 \tab 9 \tab 19 \tab 60 \tab 5 \tab 44 \tab 11 \tab 18 \tab 5 \tab 93 \tab 51 \tab 87 \tab 25 \tab 113 \tab 12 \tab 151 \tab 61 \tab 8 \tab 301 \tab 103 \tab 18 \tab 43 \tab 6 \tab 36 \tab 72 \tab 11 \tab 2 \tab 55 \tab 19 \tab 26 \tab 18 \tab 9 \tab 3 \tab 5 \tab 21 \tab 17 \tab 22 \tab 8 \tab 45 \tab 15 \tab 47 \tab 19 \tab 46 \tab 2 \tab 34 \tab 9 \tab 16 \tab 23 \tab 21 \tab 11 \tab 10 \tab 14 \tab 6 \tab 31 \tab 38 \tab 44 \tab 37 \tab 21 \tab 16 \tab 13 \tab 2 \tab 10 \tab 12 \tab 12 \tab 5 \tab 26 \tab 6 \tab 3 \tab 15 \tab 2 \tab 25 \tab 12 \tab 41 \tab 7 \tab 21 \tab 16 \tab 10 \tab 7 \tab 9 \tab 16 \tab 3 \tab 7 \tab 2 \tab 4 \tab 3 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{10} agricultural production, crops
\item \code{11} agricultural production, livestock
\item \code{20} agricultural services, except horticultural
\item \code{21} horticultural services
\item \code{30} forestry
\item \code{31} fishing, hunting, and trapping, etc.
\item \code{40} metal mining
\item \code{41} coal mining
\item \code{42} crude petroleum and natural gas extraction
\item \code{50} nonmetallic mining and quarrying, except fuel
\item \code{60} construction
\item \code{100} meat products
\item \code{101} dairy products
\item \code{102} canned and preserved fruits and vegetables
\item \code{110} grain mill products
\item \code{111} bakery products
\item \code{112} sugar and confectionery products
\item \code{120} beverage industries
\item \code{121} miscellaneous food preparations and kindred products
\item \code{122} not specified food industries
\item \code{130} tobacco manufactures
\item \code{132} knitting mills
\item \code{140} dyeing and finishing textiles, except wool and knit goods
\item \code{141} floor coverings, except hard surface
\item \code{142} yarn, thread, and fabric mills
\item \code{150} miscellaneous textile mill products
\item \code{151} apparel and accessories, except knit
\item \code{152} miscellaneous fabricated textile products
\item \code{160} pulp, paper, and paperboard mills
\item \code{161} miscellaneous paper and pulp products
\item \code{162} paperboard containers and boxes
\item \code{171} newspaper publishing and printing
\item \code{172} printing, publishing, and allied industries, except newspapers
\item \code{180} plastics, synthetics, and resins
\item \code{181} drugs
\item \code{182} soaps and cosmetics
\item \code{185} paints, varnishes, and related products
\item \code{190} agricultural chemicals
\item \code{191} industrial and miscellaneous chemicals
\item \code{192} petroleum and coal products
\item \code{200} petroleum refining
\item \code{201} miscellaneous petroleum and coal products
\item \code{210} tires and inner tubes
\item \code{211} other rubber products, and plastics footwear and belting
\item \code{212} miscellaneous plastics products
\item \code{220} leather tanning and finishing
\item \code{221} footwear, except rubber and plastic
\item \code{222} leather products, except footwear
\item \code{230} logging
\item \code{231} sawmills, planning mills, and millwork
\item \code{232} wood buildings and mobile homes
\item \code{241} miscellaneous wood products
\item \code{242} furniture and fixtures
\item \code{250} glass and glass products
\item \code{251} cement, concrete, gypsum, and plaster products
\item \code{252} structural clay products
\item \code{261} pottery and related products
\item \code{262} miscellaneous nonmetallic mineral and stone products
\item \code{270} blast furnaces, steelworks, rolling and finishing mills
\item \code{271} iron and steel foundries
\item \code{272} primary aluminum industries
\item \code{280} other primary metal industries
\item \code{281} cutlery, hand tools, and other hardware
\item \code{282} fabricated structural metal products
\item \code{290} screw machine products
\item \code{291} metal forgings and stampings
\item \code{292} ordnance
\item \code{300} miscellaneous fabricated metal products
\item \code{301} not specified metal industries
\item \code{310} engines and turbines
\item \code{311} farm machinery and equipment
\item \code{312} construction and material handling machines
\item \code{320} metalworking machinery
\item \code{321} office and accounting machines
\item \code{322} electronic computing equipment
\item \code{331} machinery, except electrical, n.e.c.
\item \code{332} not specified machinery
\item \code{340} household appliances
\item \code{341} radio, tv, and communication equipment
\item \code{342} electrical machinery, equipment, and supplies, n.e.c.
\item \code{350} not specified electrical machinery, equipment, and supplies
\item \code{351} motor vehicles and motor vehicle equipment
\item \code{352} aircraft and parts
\item \code{360} ship and boat building and repairing
\item \code{361} railroad locomotives and equipment
\item \code{362} guided missiles, space vehicles, and parts
\item \code{370} cycles and miscellaneous transportation equipment
\item \code{371} scientific and controlling instruments
\item \code{372} optical and health services supplies
\item \code{380} photographic equipment and supplies
\item \code{381} watches, clocks, and clockwork operated devices
\item \code{382} not specified professional equipment
\item \code{390} toys, amusement, and sporting goods
\item \code{391} miscellaneous manufacturing industries
\item \code{392} not specified manufacturing industries
\item \code{400} railroads
\item \code{401} bus service and urban transit
\item \code{402} taxicab service
\item \code{410} trucking service
\item \code{411} warehousing and storage
\item \code{412} u.s. postal service
\item \code{420} water transportation
\item \code{421} air transportation
\item \code{422} pipe lines, except natural gas
\item \code{432} services incidental to transportation
\item \code{440} radio and television broadcasting
\item \code{441} telephone (wire and radio)
\item \code{442} telegraph and miscellaneous communication service
\item \code{460} electric light and power
\item \code{461} gas and steam supply systems
\item \code{462} electric and gas, and other combinations
\item \code{470} water supply and irrigation
\item \code{471} sanitary services
\item \code{472} not specified utilities
\item \code{500} motor vehicles and equipment
\item \code{501} furniture and home furnishings
\item \code{502} lumber and construction materials
\item \code{510} sporting goods, toys, and hobby goods
\item \code{511} metals and minerals, except petroleum
\item \code{512} electrical goods
\item \code{521} hardware, plumbing and heating supplies
\item \code{522} not specified electrical and hardware products
\item \code{530} machinery, equipment, and supplies
\item \code{531} scrap and waste materials
\item \code{532} miscellaneous wholesale, durable goods
\item \code{540} paper and paper products
\item \code{541} drugs, chemicals, and allied products
\item \code{542} apparel, fabrics, and notions
\item \code{550} groceries and related products
\item \code{551} farm products-raw materials
\item \code{552} petroleum products
\item \code{560} alcoholic beverages
\item \code{561} farm supplies
\item \code{562} miscellaneous wholesale, nondurable goods
\item \code{571} not specified wholesale trade
\item \code{580} lumber and building material retailing
\item \code{581} hardware stores
\item \code{582} retail nurseries and garden stores
\item \code{590} mobile home dealers
\item \code{591} department stores
\item \code{592} variety stores
\item \code{600} miscellaneous general merchandise stores
\item \code{601} grocery stores
\item \code{602} dairy products stores
\item \code{610} retail bakeries
\item \code{611} food stores, n.e.c.
\item \code{612} motor vehicle dealers
\item \code{620} auto and home supply stores
\item \code{621} gasoline service stations
\item \code{622} miscellaneous vehicle dealers
\item \code{630} apparel and accessory stores, except shoe
\item \code{631} shoe stores
\item \code{632} furniture and home furnishings stores
\item \code{640} household appliances, tv, and radio stores
\item \code{641} eating and drinking places
\item \code{642} drug stores
\item \code{650} liquor stores
\item \code{651} sporting goods, bicycles, and hobby stores
\item \code{652} book and stationery stores
\item \code{660} jewelry stores
\item \code{661} sewing, needlework, and piece goods stores
\item \code{662} mail order houses
\item \code{670} vending machine operators
\item \code{671} direct selling establishments
\item \code{672} fuel and ice dealers
\item \code{681} retail florists
\item \code{682} miscellaneous retail stores
\item \code{691} not specified retail trade
\item \code{700} banking
\item \code{701} savings and loan associations
\item \code{702} credit agencies, n.e.c.
\item \code{710} security, commodity brokerage, and investment companies
\item \code{711} insurance
\item \code{712} real estate, including real estate-insurance-law offices
\item \code{721} advertising
\item \code{722} services to dwellings and other buildings
\item \code{730} commercial research, development, and testing labs
\item \code{731} personnel supply services
\item \code{732} business management and consulting services
\item \code{740} computer and data processing services
\item \code{741} detective and protective services
\item \code{742} business services, n.e.c.
\item \code{750} automotive services, except repair
\item \code{751} automotive repair shops
\item \code{752} electrical repair shops
\item \code{760} miscellaneous repair services
\item \code{761} private households
\item \code{762} hotels and motels
\item \code{770} lodging places, except hotels and motels
\item \code{771} laundry, cleaning, and garment services
\item \code{772} beauty shops
\item \code{780} barber shops
\item \code{781} funeral service and crematories
\item \code{782} shoe repair shops
\item \code{790} dressmaking shops
\item \code{791} miscellaneous personal services
\item \code{800} theaters and motion pictures
\item \code{801} bowling alleys, billiard and pool parlors
\item \code{802} miscellaneous entertainment and recreation services
\item \code{812} offices of physicians
\item \code{820} offices of dentists
\item \code{821} offices of chiropractors
\item \code{822} offices of optometrists
\item \code{830} offices of health practitioners, n.e.c.
\item \code{831} hospitals
\item \code{832} nursing and personal care facilities
\item \code{840} health services, n.e.c.
\item \code{841} legal services
\item \code{842} elementary and secondary schools
\item \code{850} colleges and universities
\item \code{851} business, trade, and vocational schools
\item \code{852} libraries
\item \code{860} educational services, n.e.c.
\item \code{861} job training and vocational rehabilitation services
\item \code{862} child day care services
\item \code{870} residential care facilities, without nursing
\item \code{871} social services, n.e.c.
\item \code{872} museums, art galleries and zoos
\item \code{880} religious organizations
\item \code{881} membership organizations
\item \code{882} engineering, architectural, and surveying services
\item \code{890} accounting, auditing, and bookkeeping services
\item \code{891} noncommercial educational and scientific research
\item \code{892} miscellaneous professional and related services
\item \code{900} executive and legislative offices
\item \code{901} general government, n.e.c.
\item \code{910} justice, public order, and safety
\item \code{921} public finance, taxation, and monetary policy
\item \code{922} administration of human resources programs
\item \code{930} administration of environmental quality and household programs
\item \code{931} administration of economic programs
\item \code{932} national security and international affairs
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
