% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_06.R
\name{spindstatus}
\alias{spindstatus}
\title{Spind10 coding status}
\source{
General Social Survey https://gss.norc.org
}
\description{
spindstatus
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab found, coded \tab found, iap \tab found, no answer \tab not found, coded \tab not found, iap \tab not found, no answer \tab not found, no code assigned \tab iap \tab Total \cr
   1972 \tab 436 \tab 255 \tab 11 \tab 446 \tab 325 \tab 22 \tab 118 \tab - \tab 1613 \cr
   1973 \tab 771 \tab 439 \tab 24 \tab 106 \tab 97 \tab 6 \tab 61 \tab - \tab 1504 \cr
   1974 \tab 746 \tab 428 \tab 43 \tab 144 \tab 95 \tab 2 \tab 26 \tab - \tab 1484 \cr
   1975 \tab 606 \tab 361 \tab 9 \tab 253 \tab 210 \tab 5 \tab 46 \tab - \tab 1490 \cr
   1976 \tab 733 \tab 488 \tab 9 \tab 114 \tab 124 \tab 1 \tab 30 \tab - \tab 1499 \cr
   1977 \tab 484 \tab 323 \tab 10 \tab 337 \tab 315 \tab - \tab 61 \tab - \tab 1530 \cr
   1978 \tab 868 \tab 639 \tab 11 \tab - \tab 14 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 721 \tab 566 \tab 6 \tab 85 \tab 74 \tab 1 \tab 15 \tab - \tab 1468 \cr
   1982 \tab 819 \tab 855 \tab 22 \tab 72 \tab 84 \tab 1 \tab 7 \tab - \tab 1860 \cr
   1983 \tab 809 \tab 650 \tab 16 \tab 59 \tab 54 \tab - \tab 11 \tab - \tab 1599 \cr
   1984 \tab 772 \tab 677 \tab 9 \tab 3 \tab 10 \tab - \tab 2 \tab - \tab 1473 \cr
   1985 \tab 686 \tab 602 \tab 8 \tab 108 \tab 107 \tab - \tab 23 \tab - \tab 1534 \cr
   1986 \tab 749 \tab 666 \tab 10 \tab 10 \tab 32 \tab - \tab 3 \tab - \tab 1470 \cr
   1987 \tab 722 \tab 777 \tab 12 \tab 107 \tab 172 \tab 5 \tab 24 \tab - \tab 1819 \cr
   1988 \tab 719 \tab 728 \tab 13 \tab 7 \tab 14 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 737 \tab 666 \tab 17 \tab 46 \tab 68 \tab 1 \tab 2 \tab - \tab 1537 \cr
   1990 \tab 669 \tab 672 \tab 17 \tab 2 \tab 12 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 749 \tab 733 \tab 17 \tab 3 \tab 15 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 788 \tab 743 \tab 14 \tab 20 \tab 38 \tab 1 \tab 2 \tab - \tab 1606 \cr
   1994 \tab 1423 \tab 1491 \tab 24 \tab 20 \tab 30 \tab 2 \tab 2 \tab - \tab 2992 \cr
   1996 \tab 1314 \tab 1551 \tab 18 \tab 7 \tab 12 \tab - \tab 2 \tab - \tab 2904 \cr
   1998 \tab 1262 \tab 1505 \tab 34 \tab 8 \tab 22 \tab 1 \tab - \tab - \tab 2832 \cr
   2000 \tab 1165 \tab 1608 \tab 44 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1185 \tab 1547 \tab 33 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1356 \tab 1406 \tab 50 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2001 \tab 2465 \tab 44 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 925 \tab 1079 \tab 19 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 834 \tab 1187 \tab 23 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 25049 \tab 25107 \tab 567 \tab 1957 \tab 1924 \tab 48 \tab 435 \tab 17303 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} found, iap
\item \code{2} found, no answer
\item \code{3} found, coded
\item \code{4} not found, iap
\item \code{5} not found, no answer
\item \code{6} not found, coded
\item \code{7} not found, no code assigned
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
