% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{spkcom}
\alias{spkcom}
\title{spkcom}
\source{
General Social Survey https://gss.norc.org
}
\description{
Allow communist to speak
}
\details{
Question 79. Now, I should like to ask you some questions about a man who admits he is a Communist: a. Suppose this admitted Communist wanted to make a speech in your community. Should he be allowed to speak, or not?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab don't know \tab no answer \tab not allowed \tab yes, allowed to speak \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 54 \tab 3 \tab 721 \tab 835 \tab - \tab - \tab 1613 \cr
   1973 \tab 35 \tab 1 \tab 570 \tab 898 \tab - \tab - \tab 1504 \cr
   1974 \tab 42 \tab 3 \tab 582 \tab 857 \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab - \tab - \tab - \tab 1490 \tab - \tab 1490 \cr
   1976 \tab 33 \tab 2 \tab 647 \tab 817 \tab - \tab - \tab 1499 \cr
   1977 \tab 31 \tab 5 \tab 648 \tab 846 \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab - \tab - \tab - \tab 1532 \tab - \tab 1532 \cr
   1980 \tab 37 \tab 1 \tab 621 \tab 809 \tab - \tab - \tab 1468 \cr
   1982 \tab 81 \tab 8 \tab 769 \tab 1002 \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab - \tab - \tab - \tab 1599 \tab - \tab 1599 \cr
   1984 \tab 36 \tab 5 \tab 562 \tab 870 \tab - \tab - \tab 1473 \cr
   1985 \tab 41 \tab 5 \tab 616 \tab 872 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab - \tab - \tab - \tab 1470 \tab - \tab 1470 \cr
   1987 \tab 43 \tab 4 \tab 706 \tab 1066 \tab - \tab - \tab 1819 \cr
   1988 \tab 23 \tab 1 \tab 365 \tab 588 \tab 504 \tab - \tab 1481 \cr
   1989 \tab 25 \tab 3 \tab 343 \tab 662 \tab 504 \tab - \tab 1537 \cr
   1990 \tab 27 \tab 3 \tab 298 \tab 589 \tab 455 \tab - \tab 1372 \cr
   1991 \tab 20 \tab 3 \tab 304 \tab 666 \tab 524 \tab - \tab 1517 \cr
   1993 \tab 28 \tab 3 \tab 308 \tab 736 \tab 531 \tab - \tab 1606 \cr
   1994 \tab 42 \tab 4 \tab 619 \tab 1331 \tab 996 \tab - \tab 2992 \cr
   1996 \tab 51 \tab 2 \tab 640 \tab 1230 \tab 981 \tab - \tab 2904 \cr
   1998 \tab 50 \tab 6 \tab 577 \tab 1249 \tab 950 \tab - \tab 2832 \cr
   2000 \tab 54 \tab 6 \tab 585 \tab 1216 \tab 956 \tab - \tab 2817 \cr
   2002 \tab 21 \tab 1 \tab 266 \tab 636 \tab 1841 \tab - \tab 2765 \cr
   2004 \tab 15 \tab 3 \tab 257 \tab 623 \tab 1914 \tab - \tab 2812 \cr
   2006 \tab 39 \tab 2 \tab 623 \tab 1339 \tab 2507 \tab - \tab 4510 \cr
   2008 \tab 34 \tab 2 \tab 442 \tab 874 \tab 671 \tab - \tab 2023 \cr
   2010 \tab 33 \tab 6 \tab 423 \tab 819 \tab 763 \tab - \tab 2044 \cr
   2012 \tab 44 \tab 7 \tab 395 \tab 862 \tab 666 \tab - \tab 1974 \cr
   2014 \tab 49 \tab 5 \tab 504 \tab 1156 \tab 824 \tab - \tab 2538 \cr
   2016 \tab 52 \tab 15 \tab 523 \tab 1300 \tab 977 \tab - \tab 2867 \cr
   2018 \tab 61 \tab 5 \tab 453 \tab 1055 \tab 774 \tab - \tab 2348 \cr
   2021 \tab 3 \tab - \tab 347 \tab 958 \tab 2714 \tab 10 \tab 4032 \cr
   2022 \tab 29 \tab 6 \tab 336 \tab 789 \tab 2380 \tab 4 \tab 3544 \cr
   Total \tab 1133 \tab 120 \tab 15050 \tab 27550 \tab 28523 \tab 14 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes, allowed to speak
\item \code{2} not allowed
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
