% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{spkhomo}
\alias{spkhomo}
\title{spkhomo}
\source{
General Social Survey https://gss.norc.org
}
\description{
Allow homosexual to speak
}
\details{
Question 81. And what about a man who admits that he is a homosexual? a. Suppose this admitted homosexual wanted to make a speech in your community. Should he be allowed to speak, or not?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no answer \tab not allowed \tab yes, allowed to speak \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 56 \tab 3 \tab 532 \tab 913 \tab - \tab 1504 \cr
   1974 \tab - \tab 67 \tab 1 \tab 491 \tab 925 \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 47 \tab 3 \tab 522 \tab 927 \tab - \tab 1499 \cr
   1977 \tab - \tab 56 \tab 2 \tab 528 \tab 944 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 46 \tab - \tab 455 \tab 967 \tab - \tab 1468 \cr
   1982 \tab - \tab 84 \tab 9 \tab 585 \tab 1182 \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 52 \tab 5 \tab 416 \tab 1000 \tab - \tab 1473 \cr
   1985 \tab - \tab 52 \tab 4 \tab 456 \tab 1022 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 49 \tab 5 \tab 544 \tab 1221 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 37 \tab 2 \tab 257 \tab 681 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 27 \tab 3 \tab 217 \tab 786 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 34 \tab 3 \tab 206 \tab 674 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 24 \tab 3 \tab 214 \tab 752 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 35 \tab 2 \tab 199 \tab 839 \tab - \tab 1606 \cr
   1994 \tab 996 \tab 47 \tab 8 \tab 363 \tab 1578 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 44 \tab 1 \tab 325 \tab 1553 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 54 \tab 3 \tab 306 \tab 1519 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 56 \tab 6 \tab 305 \tab 1494 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 11 \tab 3 \tab 138 \tab 772 \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 6 \tab 2 \tab 140 \tab 750 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 33 \tab 9 \tab 326 \tab 1635 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 26 \tab 4 \tab 225 \tab 1097 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 16 \tab 7 \tab 168 \tab 1090 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 23 \tab 8 \tab 155 \tab 1122 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 13 \tab 10 \tab 180 \tab 1511 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 15 \tab 13 \tab 195 \tab 1667 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 14 \tab 6 \tab 173 \tab 1381 \tab - \tab 2348 \cr
   2021 \tab 2714 \tab 1 \tab 1 \tab 108 \tab 1196 \tab 12 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 31300 \tab 1025 \tab 126 \tab 8729 \tab 31198 \tab 12 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes, allowed to speak
\item \code{2} not allowed
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
