% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_31.R
\name{spkmil}
\alias{spkmil}
\title{spkmil}
\source{
General Social Survey https://gss.norc.org
}
\description{
Allow militarist to speak
}
\details{
Question 80. Consider a person who advocates doing away with elections and letting the military run the country. a. If such a person wanted to make a speech in your community, should he be allowed to speak, or not?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no answer \tab not allowed \tab yes, allowed to speak \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 26 \tab 4 \tab 656 \tab 813 \tab - \tab 1499 \cr
   1977 \tab - \tab 28 \tab 1 \tab 729 \tab 772 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 30 \tab 2 \tab 602 \tab 834 \tab - \tab 1468 \cr
   1982 \tab - \tab 69 \tab 3 \tab 829 \tab 959 \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 22 \tab 6 \tab 609 \tab 836 \tab - \tab 1473 \cr
   1985 \tab - \tab 33 \tab 5 \tab 661 \tab 835 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 59 \tab 5 \tab 762 \tab 993 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 30 \tab 5 \tab 399 \tab 543 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 17 \tab 4 \tab 402 \tab 610 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 29 \tab 3 \tab 364 \tab 521 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 22 \tab 4 \tab 352 \tab 615 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 27 \tab 3 \tab 354 \tab 691 \tab - \tab 1606 \cr
   1994 \tab 996 \tab 41 \tab 3 \tab 680 \tab 1272 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 39 \tab 2 \tab 673 \tab 1209 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 43 \tab 3 \tab 589 \tab 1247 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 41 \tab 5 \tab 643 \tab 1172 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 12 \tab 1 \tab 276 \tab 635 \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 9 \tab 2 \tab 287 \tab 600 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 33 \tab 2 \tab 650 \tab 1318 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 30 \tab 3 \tab 448 \tab 871 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 17 \tab 6 \tab 393 \tab 865 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 23 \tab 3 \tab 389 \tab 893 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 18 \tab 7 \tab 495 \tab 1194 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 22 \tab 7 \tab 505 \tab 1356 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 27 \tab 4 \tab 438 \tab 1105 \tab - \tab 2348 \cr
   2021 \tab 2714 \tab - \tab - \tab 496 \tab 805 \tab 17 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 34288 \tab 747 \tab 93 \tab 13681 \tab 23564 \tab 17 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes, allowed to speak
\item \code{2} not allowed
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
