% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_05.R
\name{spoccindv}
\alias{spoccindv}
\title{spoccindv}
\source{
General Social Survey https://gss.norc.org
}
\description{
Spouse's occ/ind verbatim retrieved
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab no verbatims for sp \tab verbatims for sp \tab Total \cr
   1972 \tab 911 \tab 257 \tab 445 \tab 1613 \cr
   1973 \tab 270 \tab 440 \tab 794 \tab 1504 \cr
   1974 \tab 267 \tab 429 \tab 788 \tab 1484 \cr
   1975 \tab 514 \tab 361 \tab 615 \tab 1490 \cr
   1976 \tab 269 \tab 490 \tab 740 \tab 1499 \cr
   1977 \tab 713 \tab 324 \tab 493 \tab 1530 \cr
   1978 \tab 14 \tab 643 \tab 875 \tab 1532 \cr
   1980 \tab 175 \tab 566 \tab 727 \tab 1468 \cr
   1982 \tab 164 \tab 859 \tab 837 \tab 1860 \cr
   1983 \tab 124 \tab 650 \tab 825 \tab 1599 \cr
   1984 \tab 15 \tab 680 \tab 778 \tab 1473 \cr
   1985 \tab 238 \tab 604 \tab 692 \tab 1534 \cr
   1986 \tab 45 \tab 670 \tab 755 \tab 1470 \cr
   1987 \tab 308 \tab 775 \tab 736 \tab 1819 \cr
   1988 \tab 21 \tab 728 \tab 732 \tab 1481 \cr
   1989 \tab 117 \tab 671 \tab 749 \tab 1537 \cr
   1990 \tab 14 \tab 680 \tab 678 \tab 1372 \cr
   1991 \tab 18 \tab 737 \tab 762 \tab 1517 \cr
   1993 \tab 61 \tab 745 \tab 800 \tab 1606 \cr
   1994 \tab 54 \tab 1491 \tab 1447 \tab 2992 \cr
   1996 \tab 21 \tab 1555 \tab 1328 \tab 2904 \cr
   1998 \tab 31 \tab 1511 \tab 1290 \tab 2832 \cr
   2000 \tab - \tab 1637 \tab 1180 \tab 2817 \cr
   2002 \tab - \tab 1552 \tab 1213 \tab 2765 \cr
   2004 \tab - \tab 1441 \tab 1371 \tab 2812 \cr
   2006 \tab - \tab 2490 \tab 2020 \tab 4510 \cr
   2008 \tab - \tab 1087 \tab 936 \tab 2023 \cr
   2010 \tab - \tab 1197 \tab 847 \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab 3544 \cr
   Total \tab 21667 \tab 25270 \tab 25453 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} no verbatims for sp
\item \code{1} verbatims for sp
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
