% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_41.R
\name{spoth16}
\alias{spoth16}
\title{Other protestant denominations}
\source{
General Social Survey https://gss.norc.org
}
\description{
spoth16
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 7th day adventist \tab assembly of god \tab brethren church, brethren \tab brethren, plymouth \tab christian disciples \tab christian reform \tab christian scientist \tab christian; central christian \tab church of christ \tab church of christ, evangelical \tab church of god in christ holiness \tab churches of god(except with christ and holiness) \tab congregationalist, 1st congreg \tab covenant \tab don't know \tab dutch reform \tab eden evangelist \tab evangelical, evangelist \tab first christian \tab first reformed \tab four square gospel \tab friends \tab full gospel \tab grace reformed \tab holiness (nazarene) \tab holiness church of god \tab holiness; church of holiness \tab holy roller \tab mennonite \tab mormon \tab nazarene \tab pentecostal \tab pentecostal assembly of god \tab pentecostal holiness, holiness pentecostal \tab pilgrim holiness \tab quaker \tab reformed \tab reformed church of christ \tab salvation army \tab swedish mission \tab united brethren, united brethren in christ \tab united church of christ \tab united church, unity church \tab wesleyan \tab zion union \tab zion union apostolic \tab advent christian \tab african methodist \tab apostolic faith \tab christ adelphians \tab christian catholic \tab church of the living god \tab jehovah's witnesses \tab lds--reorganized \tab other fundamentalist \tab christian & missionary alliance \tab church of god in christ \tab evangelical reformed \tab reformed united church of christ \tab unitarian, universalist \tab united church of christianity \tab evangelist free church \tab christ church unity \tab disciples of christ \tab no answer \tab church of god, saint & christ \tab first christian disciples of christ \tab first church \tab free will baptist \tab independent \tab mission covenant \tab missionary baptist \tab moravian \tab open bible \tab pentecostal church of god \tab evangelical congregational \tab zion union apostolic--reformed \tab american reform \tab federated church \tab free methodist \tab pentecostal apostolic \tab amish \tab apostolic christian \tab christ in christian union \tab church of the first born \tab grace brethren \tab triumph church of god \tab evangelical united brethren \tab mennonite brethren \tab mind science \tab evangelical covenant \tab ind bible, bible, bible fellowship \tab united church of canada \tab missionary church \tab wesleyan methodist--pilgrim \tab brother of christ \tab carmelite \tab community church \tab hungarian reformed \tab bible missionary \tab primitive baptist \tab sanctified, sanctification \tab apostolic church \tab baptist (northern) \tab christ in god \tab evangelical methodist \tab faith christian \tab united holiness \tab calvary bible \tab independent fundamental church of america \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1370 \tab 1 \tab 7 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 8 \tab 14 \tab 1 \tab 1 \tab 7 \tab 12 \tab 1 \tab 9 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 5 \tab 3 \tab 9 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 2 \tab 4 \tab 1 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1371 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab 15 \tab 12 \tab - \tab - \tab 7 \tab 6 \tab 2 \tab 4 \tab - \tab - \tab 4 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 2 \tab 4 \tab 2 \tab 9 \tab - \tab - \tab - \tab 1 \tab 7 \tab - \tab - \tab 1 \tab 2 \tab 7 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1385 \tab 2 \tab - \tab 2 \tab - \tab - \tab 4 \tab 4 \tab 13 \tab 7 \tab - \tab - \tab 7 \tab 5 \tab 1 \tab 1 \tab 1 \tab - \tab 8 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 7 \tab - \tab 2 \tab 5 \tab 5 \tab 7 \tab 1 \tab - \tab - \tab 1 \tab 5 \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1386 \tab 5 \tab 4 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 11 \tab 19 \tab - \tab - \tab 8 \tab 8 \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 7 \tab - \tab 1 \tab - \tab 5 \tab 8 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1425 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 6 \tab 12 \tab - \tab - \tab 8 \tab 6 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 7 \tab 4 \tab 16 \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 4 \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1401 \tab 6 \tab 4 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 7 \tab 14 \tab - \tab - \tab 11 \tab 10 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 4 \tab 6 \tab 8 \tab - \tab 4 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab - \tab - \tab 6 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 6 \tab - \tab 1 \tab - \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1345 \tab 2 \tab 10 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 7 \tab 19 \tab 1 \tab 1 \tab 8 \tab 8 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 8 \tab 4 \tab 10 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1735 \tab 5 \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab 3 \tab 10 \tab 13 \tab - \tab - \tab 4 \tab 7 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 7 \tab - \tab 3 \tab 6 \tab 2 \tab 10 \tab 2 \tab 3 \tab - \tab 1 \tab 6 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 10 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1463 \tab 1 \tab 5 \tab - \tab - \tab - \tab 2 \tab - \tab 13 \tab 14 \tab - \tab 1 \tab 8 \tab 6 \tab - \tab 5 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 24 \tab 3 \tab 6 \tab - \tab 4 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 11 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1345 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 12 \tab 13 \tab - \tab 1 \tab 7 \tab 6 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 17 \tab 1 \tab 11 \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 12 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1422 \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 12 \tab 15 \tab - \tab - \tab 9 \tab 7 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 7 \tab - \tab - \tab 19 \tab 3 \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 8 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1378 \tab 2 \tab 3 \tab - \tab - \tab - \tab 2 \tab 1 \tab 7 \tab 11 \tab - \tab - \tab 4 \tab 5 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 22 \tab 2 \tab 7 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1688 \tab 6 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab - \tab 6 \tab 10 \tab - \tab - \tab 9 \tab 9 \tab - \tab - \tab 3 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab - \tab 19 \tab 5 \tab 6 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 8 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 14 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1367 \tab 3 \tab 10 \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 5 \tab - \tab - \tab 7 \tab 6 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 23 \tab 2 \tab 12 \tab - \tab 2 \tab - \tab 2 \tab 4 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1420 \tab 3 \tab 4 \tab 2 \tab - \tab - \tab 2 \tab - \tab 3 \tab 11 \tab - \tab - \tab 8 \tab 3 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 3 \tab 15 \tab 3 \tab 8 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 4 \tab 14 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1263 \tab 6 \tab 5 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 3 \tab 7 \tab 1 \tab - \tab 2 \tab 5 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 17 \tab 2 \tab 7 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 7 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1372 \cr
   1991 \tab 1391 \tab 6 \tab 2 \tab 2 \tab - \tab - \tab - \tab 5 \tab 11 \tab 10 \tab - \tab - \tab 5 \tab 4 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab 3 \tab 21 \tab 2 \tab 9 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 11 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1495 \tab 2 \tab 7 \tab 3 \tab - \tab - \tab 7 \tab 3 \tab 9 \tab 9 \tab - \tab - \tab 5 \tab 8 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 7 \tab 1 \tab 1 \tab 6 \tab 3 \tab 8 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2917 \tab - \tab 4 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 5 \tab - \tab - \tab 4 \tab 6 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab 24 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 70190 \tab 60 \tab 80 \tab 31 \tab 2 \tab 5 \tab 43 \tab 30 \tab 156 \tab 220 \tab 3 \tab 4 \tab 128 \tab 127 \tab 8 \tab 28 \tab 17 \tab 1 \tab 35 \tab 25 \tab 1 \tab 5 \tab 4 \tab 3 \tab 1 \tab 5 \tab 2 \tab 73 \tab 6 \tab 27 \tab 224 \tab 59 \tab 163 \tab 5 \tab 29 \tab 3 \tab 17 \tab 45 \tab 4 \tab 16 \tab 2 \tab 23 \tab 62 \tab 3 \tab 12 \tab 1 \tab 2 \tab 4 \tab 1 \tab 6 \tab 3 \tab 2 \tab 1 \tab 44 \tab 6 \tab 4 \tab 7 \tab 10 \tab 9 \tab 1 \tab 22 \tab 6 \tab 5 \tab 1 \tab 26 \tab 137 \tab 2 \tab 3 \tab 2 \tab 8 \tab 2 \tab 1 \tab 13 \tab 4 \tab 4 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 2 \tab 6 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} hungarian reformed
\item \code{2} evangelical congregational
\item \code{3} ind bible, bible, bible fellowship
\item \code{5} church of prophecy
\item \code{6} new testament christian
\item \code{7} church of god, saint & christ
\item \code{8} moravian
\item \code{9} christian & missionary alliance
\item \code{10} advent christian
\item \code{11} spiritualist
\item \code{12} assembly of god
\item \code{13} free methodist
\item \code{14} apostolic faith
\item \code{15} african methodist
\item \code{16} free will baptist
\item \code{17} eden evangelist
\item \code{18} holiness (nazarene)
\item \code{19} baptist (northern)
\item \code{20} brethren church, brethren
\item \code{21} witness holiness
\item \code{22} brethren, plymouth
\item \code{23} united brethren, united brethren in christ
\item \code{24} independent
\item \code{25} christian disciples
\item \code{26} christ in christian union
\item \code{27} open bible
\item \code{28} christian catholic
\item \code{29} christ church unity
\item \code{30} christ adelphians
\item \code{31} christian; central christian
\item \code{32} christian reform
\item \code{33} christian scientist
\item \code{34} church of christ, evangelical
\item \code{35} church of christ
\item \code{36} churches of god(except with christ and holiness)
\item \code{37} church of god in christ
\item \code{38} church of god in christ holiness
\item \code{39} church of the living god
\item \code{40} congregationalist, 1st congreg
\item \code{41} community church
\item \code{42} covenant
\item \code{43} dutch reform
\item \code{44} disciples of christ
\item \code{45} evangelical, evangelist
\item \code{46} evangelical reformed
\item \code{47} evangelist free church
\item \code{48} first church
\item \code{49} first christian disciples of christ
\item \code{50} first reformed
\item \code{51} first christian
\item \code{52} full gospel
\item \code{53} four square gospel
\item \code{54} friends
\item \code{55} holy roller
\item \code{56} holiness; church of holiness
\item \code{57} pilgrim holiness
\item \code{58} jehovah's witnesses
\item \code{59} lds
\item \code{60} lds--mormon
\item \code{61} lds--reorganized
\item \code{62} lds--jesus christ; church of jesus lds
\item \code{63} mennonite
\item \code{64} mormon
\item \code{65} nazarene
\item \code{66} pentecostal assembly of god
\item \code{67} pentecostal church of god
\item \code{68} pentecostal
\item \code{69} pentecostal holiness, holiness pentecostal
\item \code{70} quaker
\item \code{71} reformed
\item \code{72} reformed united church of christ
\item \code{73} reformed church of christ
\item \code{74} religious science
\item \code{75} mind science
\item \code{76} salvation army
\item \code{77} 7th day adventist
\item \code{78} sanctified, sanctification
\item \code{79} united holiness
\item \code{80} unitarian, universalist
\item \code{81} united church of christ
\item \code{82} united church, unity church
\item \code{83} wesleyan
\item \code{84} wesleyan methodist--pilgrim
\item \code{85} zion union
\item \code{86} zion union apostolic
\item \code{87} zion union apostolic--reformed
\item \code{88} disciples of god
\item \code{89} grace reformed
\item \code{90} holiness church of god
\item \code{91} evangelical covenant
\item \code{92} mission covenant
\item \code{93} missionary baptist
\item \code{94} swedish mission
\item \code{95} unity
\item \code{96} united church of christianity
\item \code{97} other fundamentalist
\item \code{98} federated church
\item \code{99} american reform
\item \code{100} grace brethren
\item \code{101} christ in god
\item \code{102} charismatic
\item \code{103} pentecostal apostolic
\item \code{104} house of prayer
\item \code{105} latvian lutheran
\item \code{106} triumph church of god
\item \code{107} apostolic christian
\item \code{108} christ cathedral of truth
\item \code{109} bible missionary
\item \code{110} calvary bible
\item \code{111} amish
\item \code{112} evangelical methodist
\item \code{113} worldwide church of god
\item \code{114} church universal and triumphant
\item \code{115} mennonite brethren
\item \code{116} church of the first born
\item \code{117} missionary church
\item \code{118} the way ministry
\item \code{119} united church of canada
\item \code{120} evangelical united brethren
\item \code{121} the church of god of prophecy
\item \code{122} chapel of faith
\item \code{123} polish national church
\item \code{124} faith gospel tabernacle
\item \code{125} christian calvary chapel
\item \code{126} carmelite
\item \code{127} church of daniel's band
\item \code{128} christian tabernacle
\item \code{129} living word
\item \code{130} true light church of christ
\item \code{131} macedonia
\item \code{132} brother of christ
\item \code{133} primitive baptist
\item \code{134} independent fundamental church of america
\item \code{135} chinese gospel church
\item \code{136} new age spirituality
\item \code{137} new song
\item \code{138} apostolic church
\item \code{139} faith christian
\item \code{140} people's church
\item \code{141} new birth christian
\item \code{142} unity school of christianity
\item \code{143} assyrian evangelist church
\item \code{144} spirit of christ
\item \code{145} church of jesus christ of the restoration
\item \code{146} laotian christian
\item \code{148} schwenkfelder
\item \code{150} zwinglian
\item \code{151} world overcomer outreach ministry
\item \code{152} course in miracles
\item \code{153} unity of the brethren
\item \code{154} spirit filled
\item \code{155} christian union
\item \code{156} church of living christ
\item \code{157} community of christ
\item \code{158} new hope christian fellowship
\item \code{159} community christian fellowship
\item \code{160} friends in christ
\item \code{161} hawaiian ohana
\item \code{162} reformed church of jesus christ of latter day saints
\item \code{163} swedenurgian/churches of the new jerusalem
\item \code{164} divine science
\item \code{165} church of the living god
\item \code{166} united christian
\item \code{167} sanctuary
\item \code{168} rain on us deliverance ministries
\item \code{169} the word church
\item \code{170} cornerstone church
\item \code{171} life sanctuary
\item \code{172} word of faith church
\item \code{173} harvest church
\item \code{174} shephard's chapel
\item \code{175} greater new testament church
\item \code{176} vineyard church
\item \code{177} real life ministries
\item \code{178} cathedral of joy
\item \code{179} great faith ministries
\item \code{180} shield of faith ministries
\item \code{181} born again
\item \code{182} alliance
\item \code{183} jacobite apostolic
\item \code{184} church of god of israel
\item \code{185} journeys
\item \code{186} national progressive baptist
\item \code{187} new apostolic
\item \code{188} metropolitan community
\item \code{189} family life church
\item \code{190} faith fellowship
\item \code{191} faith covenant
\item \code{192} new thought
\item \code{193} free spirit ministry
\item \code{194} jacobite syrian christian church
\item \code{195} the ark church
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
