% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_05.R
\name{sppres}
\alias{sppres}
\title{sppres}
\source{
General Social Survey https://gss.norc.org
}
\description{
Spouse's occupational prestige score (1970)
}
\details{
Question 6a. What kind of work (does/did) your \link{SPOUSE} normally do? That is, what (is/was) (his/her) job called? b. What (does/did) \link{SPOUSE} actually do in that job? Tell me, what (are/were) some of (his/her) main duties?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 12 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 60 \tab 61 \tab 62 \tab 63 \tab 67 \tab 68 \tab 69 \tab 71 \tab 72 \tab 76 \tab 78 \tab 82 \tab dk, na, iap \tab 53 \tab 65 \tab 70 \tab 74 \tab 66 \tab 21 \tab not available in this year \tab iap \tab Total \cr
   1972 \tab 8 \tab 4 \tab 3 \tab 18 \tab 27 \tab 30 \tab 10 \tab 18 \tab 17 \tab 10 \tab 2 \tab 14 \tab 19 \tab 17 \tab 7 \tab 23 \tab 6 \tab 13 \tab 69 \tab 35 \tab 35 \tab 15 \tab 64 \tab 15 \tab 9 \tab 24 \tab 34 \tab 55 \tab 11 \tab 14 \tab 7 \tab 33 \tab 68 \tab 27 \tab 48 \tab 10 \tab 79 \tab 15 \tab 5 \tab 3 \tab 4 \tab 7 \tab 12 \tab 2 \tab 29 \tab 5 \tab 12 \tab 10 \tab 2 \tab 3 \tab 6 \tab 2 \tab 2 \tab 1 \tab 7 \tab 2 \tab 586 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 4 \tab 1 \tab 4 \tab 12 \tab 25 \tab 18 \tab 9 \tab 15 \tab 8 \tab 9 \tab - \tab 27 \tab 22 \tab 13 \tab 12 \tab 41 \tab 4 \tab 20 \tab 66 \tab 15 \tab 30 \tab 9 \tab 56 \tab 17 \tab 5 \tab 16 \tab 29 \tab 33 \tab 17 \tab 17 \tab 8 \tab 25 \tab 45 \tab 33 \tab 40 \tab 9 \tab 77 \tab 8 \tab 6 \tab 5 \tab 3 \tab 9 \tab 8 \tab 3 \tab 31 \tab 11 \tab 16 \tab 9 \tab 5 \tab 5 \tab 9 \tab 2 \tab 2 \tab 8 \tab 10 \tab 4 \tab 554 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1 \tab 1 \tab 3 \tab 8 \tab 27 \tab 25 \tab 10 \tab 18 \tab 3 \tab 15 \tab - \tab 24 \tab 21 \tab 15 \tab 6 \tab 48 \tab 2 \tab 16 \tab 47 \tab 28 \tab 8 \tab 3 \tab 62 \tab 19 \tab 5 \tab 20 \tab 41 \tab 34 \tab 13 \tab 11 \tab 9 \tab 42 \tab 52 \tab 30 \tab 42 \tab 11 \tab 86 \tab 15 \tab 4 \tab 6 \tab 3 \tab 7 \tab 8 \tab 2 \tab 34 \tab 7 \tab 22 \tab 17 \tab 4 \tab 3 \tab 7 \tab - \tab 3 \tab 2 \tab 1 \tab 4 \tab 527 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1 \tab 3 \tab 3 \tab 9 \tab 27 \tab 24 \tab 7 \tab 11 \tab 10 \tab 8 \tab 1 \tab 14 \tab 23 \tab 20 \tab 16 \tab 13 \tab 8 \tab 19 \tab 49 \tab 37 \tab 43 \tab 5 \tab 75 \tab 17 \tab 10 \tab 19 \tab 33 \tab 35 \tab 12 \tab 6 \tab 6 \tab 30 \tab 41 \tab 28 \tab 40 \tab 8 \tab 70 \tab 18 \tab 3 \tab 1 \tab 1 \tab 6 \tab 12 \tab 2 \tab 28 \tab 8 \tab 12 \tab 15 \tab 1 \tab 3 \tab 4 \tab 1 \tab 4 \tab 2 \tab 5 \tab 7 \tab 574 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 5 \tab 2 \tab 4 \tab 17 \tab 13 \tab 19 \tab 9 \tab 18 \tab 14 \tab 14 \tab 1 \tab 22 \tab 24 \tab 15 \tab 11 \tab 21 \tab 8 \tab 13 \tab 49 \tab 21 \tab 37 \tab 8 \tab 63 \tab 25 \tab 1 \tab 21 \tab 34 \tab 42 \tab 6 \tab 6 \tab 13 \tab 29 \tab 38 \tab 28 \tab 39 \tab 6 \tab 56 \tab 10 \tab 3 \tab 1 \tab 4 \tab 5 \tab 13 \tab 1 \tab 20 \tab 6 \tab 14 \tab 17 \tab 2 \tab 4 \tab 10 \tab 1 \tab 4 \tab 3 \tab 6 \tab 3 \tab 618 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1 \tab 5 \tab 4 \tab 11 \tab 15 \tab 17 \tab 8 \tab 14 \tab 19 \tab 11 \tab 1 \tab 20 \tab 17 \tab 12 \tab 14 \tab 22 \tab 8 \tab 12 \tab 64 \tab 19 \tab 36 \tab 10 \tab 67 \tab 25 \tab 1 \tab 14 \tab 34 \tab 34 \tab 8 \tab 8 \tab 5 \tab 25 \tab 44 \tab 26 \tab 41 \tab 5 \tab 86 \tab 13 \tab 5 \tab 4 \tab 1 \tab 2 \tab 8 \tab 1 \tab 25 \tab 7 \tab 15 \tab 12 \tab 2 \tab 7 \tab 10 \tab 3 \tab 4 \tab 2 \tab 4 \tab 1 \tab 640 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 2 \tab 3 \tab 4 \tab 17 \tab 17 \tab 13 \tab 11 \tab 13 \tab 5 \tab 6 \tab 2 \tab 18 \tab 24 \tab 10 \tab 12 \tab 16 \tab 11 \tab 11 \tab 48 \tab 26 \tab 53 \tab 9 \tab 50 \tab 22 \tab 2 \tab 20 \tab 31 \tab 31 \tab 11 \tab 6 \tab 5 \tab 28 \tab 49 \tab 42 \tab 27 \tab 5 \tab 89 \tab 9 \tab 6 \tab 4 \tab 2 \tab 10 \tab 13 \tab - \tab 26 \tab 7 \tab 16 \tab 3 \tab 2 \tab 2 \tab 6 \tab 1 \tab 8 \tab 5 \tab 3 \tab 2 \tab 656 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 4 \tab 4 \tab 2 \tab 12 \tab 25 \tab 13 \tab 5 \tab 10 \tab 9 \tab 5 \tab - \tab 15 \tab 18 \tab 17 \tab 7 \tab 22 \tab 6 \tab 8 \tab 42 \tab 21 \tab 38 \tab 9 \tab 51 \tab 23 \tab 1 \tab 29 \tab 16 \tab 34 \tab 10 \tab 4 \tab 8 \tab 37 \tab 36 \tab 29 \tab 41 \tab 4 \tab 78 \tab 8 \tab 7 \tab 2 \tab 3 \tab 13 \tab 8 \tab 1 \tab 26 \tab 18 \tab 11 \tab 8 \tab 1 \tab 3 \tab 6 \tab 3 \tab 3 \tab 4 \tab 5 \tab - \tab 642 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 10 \tab 4 \tab 1 \tab 11 \tab 30 \tab 20 \tab 9 \tab 14 \tab 22 \tab 12 \tab 1 \tab 21 \tab 25 \tab 11 \tab 9 \tab 22 \tab 8 \tab 21 \tab 60 \tab 32 \tab 43 \tab 9 \tab 70 \tab 22 \tab 6 \tab 15 \tab 22 \tab 26 \tab 10 \tab 10 \tab 5 \tab 18 \tab 38 \tab 23 \tab 29 \tab 10 \tab 73 \tab 14 \tab 2 \tab 4 \tab - \tab 9 \tab 9 \tab - \tab 34 \tab 10 \tab 13 \tab 9 \tab 3 \tab 5 \tab 3 \tab 3 \tab 3 \tab 3 \tab 6 \tab 3 \tab 950 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1860 \cr
   1983 \tab 3 \tab 2 \tab 3 \tab 11 \tab 23 \tab 13 \tab 9 \tab 16 \tab 11 \tab 13 \tab 1 \tab 18 \tab 12 \tab 12 \tab 11 \tab 11 \tab 9 \tab 17 \tab 37 \tab 24 \tab 53 \tab 10 \tab 48 \tab 17 \tab 3 \tab 20 \tab 24 \tab 41 \tab 12 \tab 8 \tab 12 \tab 27 \tab 59 \tab 21 \tab 42 \tab 4 \tab 84 \tab 11 \tab 7 \tab - \tab 8 \tab 6 \tab 14 \tab 3 \tab 19 \tab 10 \tab 21 \tab 9 \tab 4 \tab 2 \tab 8 \tab 2 \tab 9 \tab 3 \tab 7 \tab 4 \tab 707 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 12 \tab 2 \tab 3 \tab 11 \tab 35 \tab 11 \tab 11 \tab 11 \tab 15 \tab 10 \tab 1 \tab 18 \tab 15 \tab 8 \tab 8 \tab 13 \tab 10 \tab 18 \tab 33 \tab 19 \tab 32 \tab 15 \tab 38 \tab 17 \tab 2 \tab 14 \tab 18 \tab 29 \tab 11 \tab 7 \tab 10 \tab 25 \tab 28 \tab 28 \tab 34 \tab 4 \tab 86 \tab 17 \tab 4 \tab - \tab 1 \tab 4 \tab 9 \tab 2 \tab 18 \tab 15 \tab 12 \tab 12 \tab 3 \tab - \tab 7 \tab 3 \tab 3 \tab 5 \tab 4 \tab - \tab 690 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 3 \tab 1 \tab 5 \tab 7 \tab 21 \tab 14 \tab 10 \tab 14 \tab 14 \tab 14 \tab 1 \tab 21 \tab 15 \tab 13 \tab 3 \tab 11 \tab 14 \tab 14 \tab 35 \tab 20 \tab 41 \tab 7 \tab 54 \tab 19 \tab 2 \tab 14 \tab 20 \tab 34 \tab 9 \tab 12 \tab 2 \tab 32 \tab 35 \tab 21 \tab 41 \tab 9 \tab 87 \tab 11 \tab 4 \tab 3 \tab 2 \tab 8 \tab 13 \tab 2 \tab 17 \tab 6 \tab 18 \tab 12 \tab 5 \tab 3 \tab 9 \tab 2 \tab 10 \tab 2 \tab 1 \tab 3 \tab 710 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 5 \tab 3 \tab 3 \tab 3 \tab 13 \tab 8 \tab 6 \tab 19 \tab 4 \tab 14 \tab 2 \tab 12 \tab 13 \tab 6 \tab 5 \tab 17 \tab 6 \tab 15 \tab 50 \tab 18 \tab 38 \tab 7 \tab 36 \tab 12 \tab 5 \tab 14 \tab 23 \tab 28 \tab 11 \tab 4 \tab 7 \tab 25 \tab 41 \tab 30 \tab 25 \tab 4 \tab 92 \tab 18 \tab 1 \tab 3 \tab 4 \tab 6 \tab 13 \tab 2 \tab 28 \tab 10 \tab 15 \tab 5 \tab 3 \tab 2 \tab 8 \tab 2 \tab 4 \tab 7 \tab 6 \tab 2 \tab 703 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1470 \cr
   1987 \tab 2 \tab 3 \tab 2 \tab 16 \tab 27 \tab 17 \tab 5 \tab 10 \tab 13 \tab 14 \tab - \tab 15 \tab 13 \tab 12 \tab 7 \tab 14 \tab 12 \tab 18 \tab 53 \tab 16 \tab 45 \tab 8 \tab 53 \tab 21 \tab 4 \tab 18 \tab 20 \tab 26 \tab 6 \tab 7 \tab 6 \tab 24 \tab 26 \tab 36 \tab 27 \tab 6 \tab 113 \tab 18 \tab 7 \tab 1 \tab 4 \tab 5 \tab 17 \tab - \tab 25 \tab 11 \tab 11 \tab 12 \tab 4 \tab 1 \tab 3 \tab 1 \tab 8 \tab 3 \tab 6 \tab 4 \tab 958 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 10 \tab 3 \tab 1 \tab 3 \tab 19 \tab 7 \tab 6 \tab 15 \tab 7 \tab 11 \tab - \tab 13 \tab 11 \tab 9 \tab 6 \tab 16 \tab 3 \tab 12 \tab 41 \tab 21 \tab 40 \tab 3 \tab 38 \tab 18 \tab 5 \tab 16 \tab 20 \tab 19 \tab 8 \tab 4 \tab 12 \tab 25 \tab 37 \tab 26 \tab 23 \tab 10 \tab 81 \tab 12 \tab 6 \tab 2 \tab 3 \tab 6 \tab 9 \tab 3 \tab 23 \tab 10 \tab 12 \tab 11 \tab 3 \tab 1 \tab 8 \tab 1 \tab 10 \tab 7 \tab 3 \tab 6 \tab 745 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 6 \tab 1 \tab 3 \tab 8 \tab 22 \tab 7 \tab 4 \tab 10 \tab 10 \tab 8 \tab 2 \tab 14 \tab 11 \tab 10 \tab 7 \tab 15 \tab 7 \tab 17 \tab 43 \tab 15 \tab 29 \tab 5 \tab 64 \tab 14 \tab 5 \tab 14 \tab 24 \tab 30 \tab 11 \tab 3 \tab 4 \tab 27 \tab 45 \tab 20 \tab 40 \tab 1 \tab 84 \tab 19 \tab 7 \tab 6 \tab 1 \tab 10 \tab 14 \tab 1 \tab 19 \tab 9 \tab 17 \tab 3 \tab 3 \tab 2 \tab 6 \tab 2 \tab 5 \tab 9 \tab 12 \tab 2 \tab 746 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1 \tab 1 \tab 3 \tab 5 \tab 21 \tab 3 \tab 3 \tab 11 \tab 9 \tab 11 \tab 2 \tab 18 \tab 11 \tab 7 \tab 6 \tab 12 \tab 2 \tab 7 \tab 34 \tab 15 \tab 26 \tab 9 \tab 45 \tab 4 \tab 3 \tab 18 \tab 23 \tab 21 \tab 4 \tab 3 \tab 10 \tab 21 \tab 31 \tab 20 \tab 33 \tab 4 \tab 75 \tab 13 \tab 6 \tab 2 \tab 3 \tab 9 \tab 13 \tab 1 \tab 25 \tab 19 \tab 12 \tab 1 \tab 1 \tab 3 \tab 8 \tab 4 \tab 2 \tab 8 \tab 10 \tab 2 \tab 695 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \tab - \tab 1606 \cr
   1994 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \tab - \tab 2992 \cr
   1996 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \tab - \tab 2832 \cr
   2000 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   Total \tab 78 \tab 43 \tab 51 \tab 179 \tab 387 \tab 259 \tab 132 \tab 237 \tab 190 \tab 185 \tab 17 \tab 304 \tab 294 \tab 207 \tab 147 \tab 337 \tab 124 \tab 251 \tab 820 \tab 382 \tab 627 \tab 141 \tab 934 \tab 307 \tab 69 \tab 306 \tab 446 \tab 552 \tab 170 \tab 130 \tab 129 \tab 473 \tab 713 \tab 468 \tab 612 \tab 110 \tab 1396 \tab 229 \tab 83 \tab 47 \tab 47 \tab 122 \tab 193 \tab 26 \tab 427 \tab 169 \tab 249 \tab 165 \tab 48 \tab 49 \tab 118 \tab 33 \tab 84 \tab 74 \tab 96 \tab 49 \tab 11701 \tab 8 \tab 2 \tab 12 \tab 20 \tab 5 \tab 2 \tab 42581 \tab 3544 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
