% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_06.R
\name{sppres10}
\alias{sppres10}
\title{sppres10}
\source{
General Social Survey https://gss.norc.org
}
\description{
Spouse occupational prestige score (2010)
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 16 \tab 17 \tab 20 \tab 21 \tab 22 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 67 \tab 68 \tab 69 \tab 70 \tab 72 \tab 73 \tab 74 \tab 75 \tab 80 \tab iap \tab 18 \tab 66 \tab 71 \tab 23 \tab 76 \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 2 \tab 1 \tab 1 \tab 6 \tab 7 \tab 24 \tab 52 \tab 7 \tab 15 \tab 26 \tab 9 \tab 3 \tab 53 \tab 34 \tab 24 \tab 5 \tab 59 \tab 31 \tab 19 \tab 40 \tab 33 \tab 17 \tab 5 \tab 11 \tab 10 \tab 21 \tab 56 \tab 45 \tab 88 \tab 18 \tab 20 \tab 22 \tab 6 \tab 1 \tab 13 \tab 5 \tab 6 \tab 1 \tab 1 \tab 1 \tab 5 \tab 16 \tab 30 \tab 7 \tab 4 \tab 22 \tab 4 \tab 1 \tab 1 \tab 7 \tab 2 \tab 7 \tab 6 \tab 5 \tab 1 \tab 2 \tab 695 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 2 \tab 5 \tab 1 \tab 3 \tab 5 \tab 19 \tab 23 \tab 3 \tab 13 \tab 22 \tab 14 \tab 5 \tab 48 \tab 49 \tab 15 \tab 6 \tab 67 \tab 24 \tab 12 \tab 46 \tab 31 \tab 14 \tab 4 \tab 21 \tab 11 \tab 19 \tab 45 \tab 29 \tab 66 \tab 26 \tab 25 \tab 29 \tab 6 \tab 3 \tab 11 \tab 3 \tab 3 \tab 6 \tab 3 \tab 2 \tab 8 \tab 9 \tab 27 \tab 6 \tab 5 \tab 28 \tab 8 \tab - \tab - \tab 12 \tab 3 \tab 9 \tab 10 \tab 6 \tab 1 \tab 4 \tab 619 \tab 3 \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab - \tab - \tab 4 \tab 4 \tab 17 \tab 37 \tab 6 \tab 10 \tab 31 \tab 3 \tab 2 \tab 56 \tab 37 \tab 16 \tab 6 \tab 48 \tab 14 \tab 22 \tab 47 \tab 26 \tab 30 \tab 2 \tab 12 \tab 12 \tab 23 \tab 54 \tab 28 \tab 69 \tab 19 \tab 24 \tab 28 \tab 8 \tab 1 \tab 15 \tab - \tab 4 \tab 1 \tab 5 \tab 3 \tab 8 \tab 18 \tab 28 \tab 6 \tab 11 \tab 36 \tab 8 \tab - \tab - \tab 10 \tab 5 \tab 4 \tab 4 \tab 1 \tab - \tab 4 \tab 607 \tab - \tab 8 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1 \tab - \tab - \tab 3 \tab 6 \tab 22 \tab 31 \tab 2 \tab 18 \tab 27 \tab 10 \tab 1 \tab 40 \tab 33 \tab 19 \tab 8 \tab 59 \tab 19 \tab 20 \tab 50 \tab 19 \tab 25 \tab 4 \tab 9 \tab 14 \tab 27 \tab 46 \tab 31 \tab 60 \tab 22 \tab 24 \tab 26 \tab 7 \tab 4 \tab 11 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 6 \tab 13 \tab 28 \tab 7 \tab 5 \tab 31 \tab 5 \tab 1 \tab - \tab 7 \tab 2 \tab 6 \tab 4 \tab 5 \tab 1 \tab 7 \tab 653 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab - \tab 2 \tab 1 \tab 4 \tab 27 \tab 24 \tab 3 \tab 14 \tab 20 \tab 10 \tab 3 \tab 51 \tab 41 \tab 19 \tab 11 \tab 60 \tab 21 \tab 17 \tab 55 \tab 33 \tab 29 \tab 2 \tab 16 \tab 5 \tab 29 \tab 39 \tab 42 \tab 56 \tab 23 \tab 17 \tab 23 \tab 3 \tab 4 \tab 11 \tab 2 \tab 7 \tab - \tab 1 \tab 1 \tab 10 \tab 11 \tab 21 \tab 9 \tab 7 \tab 24 \tab 5 \tab 1 \tab - \tab 6 \tab 5 \tab 5 \tab 6 \tab 6 \tab - \tab 2 \tab 646 \tab - \tab 8 \tab 1 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 2 \tab 1 \tab 1 \tab 5 \tab 4 \tab 15 \tab 32 \tab 5 \tab 7 \tab 16 \tab 1 \tab 4 \tab 42 \tab 32 \tab 19 \tab 12 \tab 66 \tab 16 \tab 10 \tab 44 \tab 29 \tab 18 \tab - \tab 21 \tab 6 \tab 19 \tab 50 \tab 30 \tab 57 \tab 28 \tab 23 \tab 22 \tab 6 \tab 1 \tab 17 \tab 4 \tab 1 \tab 1 \tab 2 \tab 4 \tab 6 \tab 12 \tab 21 \tab 7 \tab 5 \tab 27 \tab 11 \tab - \tab - \tab 5 \tab 2 \tab 7 \tab 9 \tab 4 \tab - \tab 1 \tab 733 \tab 1 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1 \tab 1 \tab 1 \tab 8 \tab 6 \tab 31 \tab 19 \tab - \tab 12 \tab 16 \tab 3 \tab 5 \tab 43 \tab 32 \tab 11 \tab 5 \tab 62 \tab 24 \tab 13 \tab 59 \tab 27 \tab 25 \tab 4 \tab 20 \tab 17 \tab 24 \tab 47 \tab 26 \tab 63 \tab 32 \tab 25 \tab 29 \tab 16 \tab 7 \tab 13 \tab 3 \tab 2 \tab 3 \tab 3 \tab 3 \tab 10 \tab 9 \tab 19 \tab 9 \tab 18 \tab 19 \tab 6 \tab 3 \tab - \tab 8 \tab 3 \tab 6 \tab 3 \tab 6 \tab - \tab 1 \tab 660 \tab 2 \tab 6 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 5 \tab 3 \tab - \tab 3 \tab 4 \tab 18 \tab 22 \tab 1 \tab 14 \tab 10 \tab 9 \tab 2 \tab 34 \tab 25 \tab 18 \tab 2 \tab 66 \tab 13 \tab 11 \tab 36 \tab 41 \tab 28 \tab 4 \tab 19 \tab 14 \tab 10 \tab 42 \tab 32 \tab 53 \tab 25 \tab 19 \tab 35 \tab 4 \tab 6 \tab 9 \tab 4 \tab 4 \tab 2 \tab 3 \tab 5 \tab 8 \tab 11 \tab 23 \tab 10 \tab 6 \tab 22 \tab 4 \tab 1 \tab - \tab 9 \tab 3 \tab 7 \tab 4 \tab 6 \tab 1 \tab - \tab 686 \tab 2 \tab 7 \tab 3 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1 \tab 3 \tab 1 \tab 3 \tab 9 \tab 27 \tab 31 \tab 8 \tab 17 \tab 25 \tab 10 \tab 5 \tab 41 \tab 32 \tab 10 \tab 15 \tab 77 \tab 22 \tab 12 \tab 62 \tab 29 \tab 22 \tab 3 \tab 22 \tab 20 \tab 21 \tab 53 \tab 34 \tab 54 \tab 30 \tab 19 \tab 17 \tab 9 \tab 5 \tab 21 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 11 \tab 12 \tab 29 \tab 4 \tab 8 \tab 19 \tab 6 \tab - \tab - \tab 5 \tab 4 \tab 2 \tab 6 \tab 8 \tab - \tab 2 \tab 962 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1 \tab 1 \tab 1 \tab 5 \tab 6 \tab 23 \tab 26 \tab 3 \tab 13 \tab 25 \tab 4 \tab 2 \tab 42 \tab 31 \tab 12 \tab 10 \tab 57 \tab 24 \tab 14 \tab 54 \tab 33 \tab 20 \tab 4 \tab 15 \tab 8 \tab 16 \tab 55 \tab 34 \tab 73 \tab 27 \tab 25 \tab 19 \tab 14 \tab 3 \tab 23 \tab 3 \tab 4 \tab 4 \tab 2 \tab 2 \tab 6 \tab 16 \tab 12 \tab 4 \tab 8 \tab 30 \tab 7 \tab 1 \tab - \tab 9 \tab 5 \tab 8 \tab 7 \tab 9 \tab 1 \tab 4 \tab 722 \tab 1 \tab 9 \tab 2 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1 \tab 4 \tab 1 \tab 3 \tab 8 \tab 17 \tab 27 \tab 7 \tab 12 \tab 18 \tab 6 \tab 4 \tab 44 \tab 28 \tab 15 \tab 8 \tab 47 \tab 11 \tab 13 \tab 52 \tab 33 \tab 13 \tab 1 \tab 18 \tab 4 \tab 13 \tab 40 \tab 38 \tab 46 \tab 28 \tab 16 \tab 29 \tab 9 \tab 8 \tab 21 \tab 1 \tab 2 \tab 4 \tab 3 \tab 1 \tab 9 \tab 12 \tab 14 \tab 18 \tab 4 \tab 22 \tab 6 \tab 1 \tab - \tab 10 \tab 3 \tab 5 \tab 11 \tab 3 \tab 1 \tab - \tab 695 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 2 \tab - \tab 1 \tab 3 \tab 4 \tab 17 \tab 30 \tab 3 \tab 14 \tab 12 \tab 4 \tab 6 \tab 53 \tab 26 \tab 16 \tab 7 \tab 62 \tab 18 \tab 13 \tab 42 \tab 33 \tab 13 \tab 4 \tab 9 \tab 9 \tab 11 \tab 58 \tab 32 \tab 56 \tab 35 \tab 17 \tab 28 \tab 7 \tab 1 \tab 17 \tab 2 \tab 3 \tab 2 \tab 1 \tab 3 \tab 11 \tab 13 \tab 19 \tab 5 \tab 8 \tab 21 \tab 6 \tab - \tab - \tab 6 \tab 2 \tab 7 \tab 6 \tab 4 \tab - \tab 3 \tab 739 \tab 1 \tab 5 \tab 4 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 3 \tab 1 \tab - \tab 2 \tab 6 \tab 11 \tab 27 \tab 3 \tab 10 \tab 19 \tab 7 \tab 3 \tab 46 \tab 16 \tab 11 \tab 3 \tab 55 \tab 13 \tab 10 \tab 55 \tab 18 \tab 13 \tab 5 \tab 13 \tab 9 \tab 20 \tab 36 \tab 37 \tab 62 \tab 28 \tab 14 \tab 23 \tab 8 \tab 4 \tab 13 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 13 \tab 15 \tab 28 \tab 7 \tab 3 \tab 25 \tab 9 \tab - \tab - \tab 8 \tab 2 \tab 4 \tab 10 \tab 2 \tab - \tab 4 \tab 712 \tab 4 \tab 8 \tab 4 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 23 \tab 28 \tab 1 \tab 13 \tab 26 \tab 6 \tab 3 \tab 49 \tab 24 \tab 17 \tab 9 \tab 76 \tab 18 \tab 18 \tab 51 \tab 33 \tab 21 \tab 3 \tab 12 \tab 7 \tab 16 \tab 32 \tab 32 \tab 42 \tab 37 \tab 16 \tab 22 \tab 14 \tab 6 \tab 18 \tab 3 \tab 7 \tab 5 \tab 1 \tab 2 \tab 3 \tab 13 \tab 23 \tab 7 \tab 6 \tab 23 \tab 6 \tab - \tab - \tab 3 \tab 2 \tab 14 \tab - \tab 5 \tab 2 \tab 4 \tab 1005 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1 \tab 2 \tab 1 \tab 4 \tab 3 \tab 14 \tab 26 \tab 1 \tab 4 \tab 10 \tab 5 \tab 3 \tab 43 \tab 20 \tab 8 \tab 6 \tab 64 \tab 17 \tab 13 \tab 44 \tab 30 \tab 19 \tab 4 \tab 9 \tab 11 \tab 19 \tab 38 \tab 23 \tab 48 \tab 28 \tab 33 \tab 14 \tab 6 \tab 5 \tab 16 \tab 1 \tab 5 \tab 5 \tab 3 \tab 3 \tab 8 \tab 13 \tab 18 \tab 4 \tab 8 \tab 27 \tab 2 \tab 2 \tab - \tab 7 \tab 4 \tab 5 \tab 4 \tab 4 \tab 3 \tab 5 \tab 751 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 17 \tab 13 \tab 2 \tab 10 \tab 27 \tab 5 \tab 2 \tab 36 \tab 30 \tab 13 \tab 7 \tab 75 \tab 14 \tab 16 \tab 49 \tab 34 \tab 16 \tab 6 \tab 8 \tab 9 \tab 16 \tab 38 \tab 36 \tab 58 \tab 30 \tab 11 \tab 19 \tab 8 \tab 5 \tab 25 \tab 6 \tab 4 \tab 5 \tab 2 \tab 3 \tab 4 \tab 6 \tab 16 \tab 11 \tab 14 \tab 25 \tab 7 \tab - \tab - \tab 12 \tab 3 \tab 6 \tab 4 \tab 8 \tab - \tab 4 \tab 751 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab 11 \tab 14 \tab 1 \tab 6 \tab 16 \tab 2 \tab 3 \tab 39 \tab 20 \tab 9 \tab 4 \tab 43 \tab 13 \tab 7 \tab 40 \tab 29 \tab 7 \tab 4 \tab 15 \tab 10 \tab 34 \tab 34 \tab 27 \tab 42 \tab 26 \tab 29 \tab 18 \tab 6 \tab 4 \tab 21 \tab 4 \tab 4 \tab 3 \tab - \tab 2 \tab 8 \tab 11 \tab 18 \tab 5 \tab 6 \tab 14 \tab 7 \tab 2 \tab - \tab 7 \tab 4 \tab 13 \tab 4 \tab 9 \tab 1 \tab 2 \tab 695 \tab - \tab 10 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1372 \cr
   1991 \tab 1 \tab 3 \tab 2 \tab 2 \tab 3 \tab 15 \tab 28 \tab 3 \tab 3 \tab 22 \tab 6 \tab 1 \tab 43 \tab 21 \tab 15 \tab 6 \tab 47 \tab 20 \tab 9 \tab 53 \tab 38 \tab 16 \tab 3 \tab 14 \tab 16 \tab 18 \tab 38 \tab 24 \tab 48 \tab 21 \tab 11 \tab 26 \tab 6 \tab 2 \tab 23 \tab 4 \tab 3 \tab 2 \tab 4 \tab 3 \tab 7 \tab 16 \tab 20 \tab 4 \tab 12 \tab 34 \tab 4 \tab 2 \tab - \tab 7 \tab 2 \tab 11 \tab 2 \tab 5 \tab 2 \tab 4 \tab 757 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 4 \tab 4 \tab 1 \tab 1 \tab 5 \tab 14 \tab 16 \tab 3 \tab 7 \tab 19 \tab 8 \tab 1 \tab 38 \tab 29 \tab 11 \tab 4 \tab 62 \tab 11 \tab 15 \tab 49 \tab 32 \tab 11 \tab 5 \tab 17 \tab 17 \tab 18 \tab 43 \tab 30 \tab 57 \tab 32 \tab 17 \tab 19 \tab 6 \tab 11 \tab 23 \tab 2 \tab 12 \tab 1 \tab 4 \tab 5 \tab 13 \tab 15 \tab 15 \tab 5 \tab 11 \tab 26 \tab 11 \tab - \tab - \tab 13 \tab 4 \tab 16 \tab 4 \tab 4 \tab - \tab 3 \tab 792 \tab - \tab 6 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2 \tab 1 \tab - \tab 8 \tab 6 \tab 26 \tab 40 \tab 6 \tab 16 \tab 27 \tab 13 \tab 2 \tab 75 \tab 38 \tab 24 \tab 11 \tab 100 \tab 24 \tab 24 \tab 92 \tab 56 \tab 27 \tab 7 \tab 23 \tab 21 \tab 25 \tab 101 \tab 46 \tab 92 \tab 50 \tab 46 \tab 45 \tab 18 \tab 8 \tab 40 \tab 11 \tab 6 \tab 6 \tab 3 \tab 9 \tab 17 \tab 23 \tab 33 \tab 12 \tab 20 \tab 55 \tab 19 \tab 1 \tab 2 \tab 14 \tab 6 \tab 29 \tab 18 \tab 12 \tab 7 \tab 3 \tab 1533 \tab 3 \tab 7 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2 \tab 4 \tab - \tab 11 \tab 11 \tab 25 \tab 34 \tab 2 \tab 22 \tab 23 \tab 8 \tab 2 \tab 59 \tab 31 \tab 32 \tab 9 \tab 103 \tab 22 \tab 22 \tab 90 \tab 52 \tab 34 \tab 9 \tab 13 \tab 25 \tab 18 \tab 74 \tab 34 \tab 65 \tab 42 \tab 35 \tab 41 \tab 17 \tab 10 \tab 32 \tab 4 \tab 4 \tab 6 \tab 3 \tab 6 \tab 17 \tab 31 \tab 42 \tab 8 \tab 18 \tab 41 \tab 19 \tab 1 \tab 1 \tab 20 \tab 9 \tab 21 \tab 18 \tab 7 \tab 4 \tab 12 \tab 1580 \tab 3 \tab 13 \tab 1 \tab 2 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1 \tab 3 \tab - \tab 6 \tab 5 \tab 18 \tab 27 \tab 4 \tab 14 \tab 23 \tab 10 \tab 3 \tab 64 \tab 23 \tab 17 \tab 11 \tab 96 \tab 29 \tab 20 \tab 79 \tab 55 \tab 32 \tab 12 \tab 21 \tab 20 \tab 25 \tab 61 \tab 33 \tab 77 \tab 51 \tab 45 \tab 46 \tab 17 \tab 8 \tab 41 \tab 10 \tab 13 \tab 6 \tab 1 \tab 5 \tab 15 \tab 38 \tab 31 \tab 3 \tab 16 \tab 37 \tab 21 \tab 1 \tab - \tab 14 \tab 6 \tab 16 \tab 18 \tab 14 \tab 2 \tab 6 \tab 1549 \tab - \tab 6 \tab 5 \tab 2 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 3 \tab 4 \tab 1 \tab - \tab 3 \tab 20 \tab 38 \tab 2 \tab 15 \tab 30 \tab 8 \tab 7 \tab 59 \tab 31 \tab 20 \tab 8 \tab 92 \tab 21 \tab 15 \tab 78 \tab 48 \tab 20 \tab 8 \tab 15 \tab 15 \tab 26 \tab 52 \tab 36 \tab 64 \tab 34 \tab 28 \tab 45 \tab 21 \tab 15 \tab 39 \tab 10 \tab 4 \tab 5 \tab 1 \tab 6 \tab 15 \tab 29 \tab 37 \tab 5 \tab 13 \tab 31 \tab 15 \tab 3 \tab 2 \tab 17 \tab 6 \tab 15 \tab 6 \tab 8 \tab 3 \tab 7 \tab 1648 \tab 5 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 5 \tab 5 \tab 1 \tab 8 \tab 5 \tab 15 \tab 36 \tab 5 \tab 9 \tab 30 \tab 6 \tab 6 \tab 60 \tab 30 \tab 20 \tab 14 \tab 89 \tab 21 \tab 10 \tab 60 \tab 52 \tab 26 \tab 8 \tab 19 \tab 17 \tab 23 \tab 58 \tab 43 \tab 69 \tab 50 \tab 43 \tab 30 \tab 16 \tab 11 \tab 33 \tab 10 \tab 12 \tab 1 \tab 3 \tab 6 \tab 19 \tab 35 \tab 34 \tab 6 \tab 8 \tab 35 \tab 15 \tab 1 \tab 1 \tab 17 \tab 8 \tab 16 \tab 9 \tab 9 \tab 1 \tab 9 \tab 1565 \tab 2 \tab 5 \tab 4 \tab 1 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1 \tab 4 \tab - \tab 5 \tab 6 \tab 23 \tab 33 \tab 4 \tab 6 \tab 28 \tab 6 \tab 4 \tab 57 \tab 25 \tab 18 \tab 14 \tab 96 \tab 30 \tab 26 \tab 90 \tab 63 \tab 21 \tab 11 \tab 28 \tab 22 \tab 31 \tab 63 \tab 41 \tab 61 \tab 54 \tab 49 \tab 47 \tab 16 \tab 8 \tab 62 \tab 10 \tab 6 \tab 2 \tab 3 \tab 6 \tab 18 \tab 38 \tab 54 \tab 8 \tab 13 \tab 41 \tab 14 \tab 2 \tab 1 \tab 17 \tab 9 \tab 13 \tab 13 \tab 13 \tab - \tab 12 \tab 1448 \tab 3 \tab 9 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 2812 \cr
   2006 \tab 7 \tab 4 \tab 1 \tab 9 \tab 12 \tab 34 \tab 50 \tab 8 \tab 16 \tab 55 \tab 10 \tab 8 \tab 76 \tab 52 \tab 48 \tab 19 \tab 164 \tab 40 \tab 23 \tab 111 \tab 66 \tab 33 \tab 15 \tab 37 \tab 31 \tab 48 \tab 93 \tab 46 \tab 89 \tab 87 \tab 47 \tab 57 \tab 24 \tab 17 \tab 64 \tab 12 \tab 18 \tab 11 \tab 6 \tab 10 \tab 27 \tab 59 \tab 84 \tab 15 \tab 28 \tab 63 \tab 29 \tab 2 \tab 3 \tab 11 \tab 13 \tab 22 \tab 15 \tab 22 \tab 7 \tab 14 \tab 2508 \tab 3 \tab 16 \tab 7 \tab 3 \tab 1 \tab - \tab - \tab 4510 \cr
   2008 \tab 1 \tab 5 \tab - \tab 5 \tab 2 \tab 16 \tab 27 \tab 3 \tab 10 \tab 21 \tab 4 \tab 2 \tab 44 \tab 15 \tab 16 \tab 5 \tab 69 \tab 10 \tab 11 \tab 58 \tab 38 \tab 16 \tab 3 \tab 11 \tab 15 \tab 19 \tab 35 \tab 23 \tab 52 \tab 50 \tab 24 \tab 24 \tab 20 \tab 14 \tab 34 \tab 7 \tab 7 \tab 9 \tab 4 \tab 1 \tab 12 \tab 31 \tab 24 \tab 11 \tab 13 \tab 33 \tab 8 \tab 2 \tab - \tab 9 \tab 5 \tab 13 \tab 7 \tab 14 \tab 2 \tab 2 \tab 1095 \tab 1 \tab 8 \tab 3 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2 \tab 2 \tab 1 \tab 3 \tab 4 \tab 24 \tab 17 \tab 4 \tab 11 \tab 25 \tab 4 \tab 2 \tab 34 \tab 25 \tab 15 \tab 2 \tab 57 \tab 14 \tab 12 \tab 44 \tab 32 \tab 18 \tab 6 \tab 10 \tab 14 \tab 15 \tab 34 \tab 26 \tab 43 \tab 43 \tab 24 \tab 25 \tab 15 \tab 8 \tab 31 \tab 6 \tab 12 \tab 6 \tab 5 \tab 5 \tab 10 \tab 18 \tab 21 \tab 6 \tab 9 \tab 29 \tab 13 \tab - \tab - \tab 9 \tab 4 \tab 9 \tab 9 \tab 6 \tab 1 \tab 8 \tab 1205 \tab 1 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 2 \tab - \tab 3 \tab 4 \tab 17 \tab 20 \tab 1 \tab 3 \tab 23 \tab - \tab 2 \tab 33 \tab 13 \tab 17 \tab 8 \tab 60 \tab 20 \tab 10 \tab 37 \tab 46 \tab 9 \tab 10 \tab 13 \tab 10 \tab 9 \tab 37 \tab 14 \tab 40 \tab 26 \tab 21 \tab 17 \tab 16 \tab 9 \tab 42 \tab 3 \tab 9 \tab 1 \tab 7 \tab 1 \tab 11 \tab 23 \tab 24 \tab - \tab 27 \tab 19 \tab 16 \tab 2 \tab 1 \tab 11 \tab 13 \tab 15 \tab 20 \tab 11 \tab 1 \tab 9 \tab 1149 \tab 3 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 1 \tab 1 \tab 8 \tab 4 \tab 21 \tab 31 \tab 2 \tab 5 \tab 24 \tab 10 \tab 2 \tab 49 \tab 23 \tab 21 \tab 7 \tab 74 \tab 24 \tab 16 \tab 62 \tab 32 \tab 15 \tab 8 \tab 27 \tab 17 \tab 19 \tab 50 \tab 27 \tab 49 \tab 49 \tab 20 \tab 26 \tab 14 \tab 8 \tab 44 \tab 15 \tab 13 \tab 3 \tab 10 \tab 2 \tab 13 \tab 38 \tab 30 \tab 15 \tab 14 \tab 41 \tab 13 \tab 2 \tab - \tab 14 \tab 7 \tab 13 \tab 10 \tab 11 \tab 2 \tab 12 \tab 1455 \tab 1 \tab 8 \tab 6 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2 \tab 4 \tab 2 \tab 7 \tab 2 \tab 28 \tab 32 \tab 7 \tab 7 \tab 33 \tab 3 \tab 6 \tab 40 \tab 28 \tab 26 \tab 5 \tab 63 \tab 23 \tab 14 \tab 61 \tab 52 \tab 22 \tab 7 \tab 24 \tab 17 \tab 27 \tab 53 \tab 33 \tab 44 \tab 52 \tab 27 \tab 29 \tab 21 \tab 9 \tab 35 \tab 8 \tab 5 \tab 3 \tab 3 \tab 11 \tab 19 \tab 27 \tab 44 \tab 12 \tab 7 \tab 56 \tab 9 \tab 1 \tab 1 \tab 23 \tab 6 \tab 15 \tab 7 \tab 16 \tab 6 \tab 5 \tab 1723 \tab - \tab 11 \tab 3 \tab 1 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 4 \tab 3 \tab - \tab 9 \tab 10 \tab 31 \tab 21 \tab 8 \tab 9 \tab 32 \tab 4 \tab 2 \tab 36 \tab 27 \tab 24 \tab 11 \tab 87 \tab 20 \tab 17 \tab 63 \tab 49 \tab 25 \tab 4 \tab 16 \tab 23 \tab 28 \tab 36 \tab 38 \tab 47 \tab 50 \tab 39 \tab 28 \tab 20 \tab 12 \tab 41 \tab 11 \tab 14 \tab 9 \tab 5 \tab 5 \tab 18 \tab 28 \tab 33 \tab 7 \tab 17 \tab 53 \tab 9 \tab 2 \tab 2 \tab 13 \tab 13 \tab 12 \tab 9 \tab 14 \tab 2 \tab 7 \tab 1179 \tab 1 \tab 7 \tab 2 \tab 2 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 2 \tab 7 \tab 1 \tab 13 \tab 5 \tab 26 \tab 40 \tab 6 \tab 20 \tab 43 \tab 19 \tab 8 \tab 75 \tab 40 \tab 20 \tab 8 \tab 118 \tab 32 \tab 19 \tab 82 \tab 81 \tab 29 \tab 10 \tab 32 \tab 33 \tab 44 \tab 62 \tab 46 \tab 76 \tab 64 \tab 47 \tab 30 \tab 45 \tab 17 \tab 71 \tab 15 \tab 14 \tab 17 \tab 8 \tab 6 \tab 40 \tab 87 \tab 98 \tab 33 \tab 23 \tab 71 \tab 32 \tab 2 \tab 4 \tab 22 \tab 15 \tab 26 \tab 10 \tab 29 \tab 1 \tab 13 \tab 2173 \tab - \tab 13 \tab 8 \tab 1 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 2 \tab 4 \tab - \tab 3 \tab 10 \tab 31 \tab 38 \tab 8 \tab 13 \tab 39 \tab 9 \tab 7 \tab 70 \tab 31 \tab 32 \tab 17 \tab 102 \tab 21 \tab 24 \tab 74 \tab 85 \tab 15 \tab 11 \tab 28 \tab 33 \tab 32 \tab 66 \tab 38 \tab 56 \tab 79 \tab 39 \tab 45 \tab 35 \tab 14 \tab 56 \tab 13 \tab 13 \tab 3 \tab 7 \tab 11 \tab 38 \tab 60 \tab 64 \tab 30 \tab 20 \tab 61 \tab 16 \tab 5 \tab - \tab 16 \tab 18 \tab 16 \tab 14 \tab 14 \tab 3 \tab 18 \tab 1891 \tab 7 \tab 14 \tab 6 \tab 1 \tab - \tab 11 \tab 7 \tab 3544 \cr
   Total \tab 62 \tab 86 \tab 24 \tab 161 \tab 182 \tab 717 \tab 990 \tab 132 \tab 388 \tab 843 \tab 236 \tab 121 \tab 1672 \tab 992 \tab 628 \tab 283 \tab 2522 \tab 694 \tab 527 \tab 2009 \tab 1388 \tab 699 \tab 196 \tab 598 \tab 522 \tab 763 \tab 1719 \tab 1134 \tab 2022 \tab 1296 \tab 919 \tab 983 \tab 464 \tab 249 \tab 986 \tab 196 \tab 224 \tab 138 \tab 110 \tab 139 \tab 443 \tab 806 \tab 1062 \tab 306 \tab 395 \tab 1141 \tab 370 \tab 42 \tab 19 \tab 378 \tab 198 \tab 393 \tab 287 \tab 300 \tab 56 \tab 193 \tab 37581 \tab 51 \tab 227 \tab 81 \tab 26 \tab 3 \tab 11 \tab 7 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
