% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_06.R
\name{sppres105plus}
\alias{sppres105plus}
\title{sppres105plus}
\source{
General Social Survey https://gss.norc.org
}
\description{
Spouse occupational prestige score using threshold method (2010)
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 58 \tab 59 \tab 6 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 68 \tab 69 \tab 7 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 76 \tab 78 \tab 79 \tab 8 \tab 80 \tab 81 \tab 83 \tab 84 \tab 85 \tab 87 \tab 88 \tab 9 \tab 90 \tab 91 \tab 92 \tab 93 \tab 94 \tab 95 \tab iap \tab 30 \tab 86 \tab 89 \tab 97 \tab 67 \tab 77 \tab 82 \tab 57 \tab not available in this release \tab Total \cr
   1972 \tab 2 \tab 5 \tab 12 \tab 15 \tab 26 \tab 20 \tab 38 \tab 2 \tab 38 \tab 12 \tab 8 \tab 15 \tab 8 \tab 15 \tab 18 \tab 8 \tab 31 \tab 23 \tab 21 \tab 54 \tab 3 \tab 6 \tab 8 \tab 5 \tab 10 \tab 2 \tab 8 \tab 15 \tab 13 \tab 9 \tab 1 \tab 8 \tab 3 \tab 7 \tab 4 \tab 11 \tab 3 \tab 30 \tab 18 \tab 2 \tab 28 \tab 7 \tab 2 \tab 7 \tab 16 \tab 77 \tab 3 \tab 34 \tab 6 \tab 1 \tab 2 \tab 6 \tab 17 \tab 8 \tab 3 \tab 7 \tab 3 \tab 4 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 9 \tab 1 \tab 1 \tab 7 \tab 7 \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab 3 \tab 39 \tab 12 \tab 10 \tab 3 \tab 5 \tab 14 \tab 4 \tab 1 \tab 10 \tab 4 \tab 695 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 5 \tab 10 \tab 5 \tab 13 \tab 14 \tab 37 \tab 1 \tab 42 \tab 11 \tab 7 \tab 15 \tab 26 \tab 8 \tab 11 \tab 3 \tab 25 \tab 20 \tab 15 \tab 57 \tab 8 \tab 8 \tab 5 \tab 1 \tab 12 \tab 1 \tab 10 \tab 13 \tab 16 \tab 9 \tab 1 \tab 10 \tab 7 \tab 4 \tab 3 \tab 10 \tab 3 \tab 17 \tab 15 \tab 2 \tab 38 \tab 4 \tab 2 \tab 5 \tab 19 \tab 54 \tab 4 \tab 20 \tab 9 \tab 5 \tab 1 \tab 9 \tab 17 \tab 14 \tab 8 \tab 9 \tab 5 \tab 6 \tab 4 \tab 1 \tab - \tab 1 \tab 2 \tab 4 \tab 2 \tab 4 \tab 5 \tab 5 \tab - \tab 3 \tab 9 \tab 3 \tab - \tab 9 \tab 30 \tab 14 \tab 7 \tab 3 \tab 11 \tab 19 \tab 3 \tab - \tab 14 \tab 8 \tab 619 \tab 1 \tab 5 \tab 8 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 4 \tab 3 \tab 24 \tab 7 \tab 12 \tab 9 \tab 38 \tab - \tab 37 \tab 11 \tab 6 \tab 10 \tab 18 \tab 9 \tab 14 \tab 5 \tab 24 \tab 13 \tab 17 \tab 50 \tab 1 \tab 10 \tab 11 \tab 4 \tab 13 \tab 3 \tab 16 \tab 18 \tab 19 \tab 2 \tab - \tab 4 \tab 5 \tab 3 \tab 7 \tab 7 \tab 5 \tab 22 \tab 20 \tab - \tab 28 \tab 6 \tab 3 \tab 9 \tab 17 \tab 64 \tab 2 \tab 20 \tab 2 \tab - \tab - \tab 7 \tab 24 \tab 12 \tab 6 \tab 6 \tab 6 \tab 7 \tab 3 \tab - \tab - \tab - \tab - \tab 6 \tab - \tab 1 \tab 5 \tab 9 \tab 1 \tab 3 \tab 7 \tab 1 \tab 3 \tab 8 \tab 37 \tab 23 \tab 9 \tab 1 \tab 6 \tab 25 \tab 1 \tab - \tab 11 \tab 4 \tab 607 \tab 1 \tab 4 \tab 3 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1484 \cr
   1975 \tab 1 \tab 2 \tab 21 \tab 8 \tab 15 \tab 11 \tab 31 \tab 1 \tab 28 \tab 14 \tab 4 \tab 16 \tab 13 \tab 13 \tab 14 \tab 5 \tab 32 \tab 11 \tab 16 \tab 55 \tab 4 \tab 8 \tab 11 \tab 3 \tab 9 \tab 3 \tab 10 \tab 16 \tab 21 \tab 3 \tab - \tab 11 \tab 6 \tab 5 \tab 4 \tab 3 \tab 1 \tab 24 \tab 28 \tab 1 \tab 39 \tab 3 \tab 3 \tab 4 \tab 15 \tab 49 \tab 2 \tab 23 \tab 8 \tab - \tab - \tab 9 \tab 15 \tab 9 \tab 2 \tab 7 \tab 2 \tab 2 \tab 10 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 2 \tab 9 \tab 7 \tab - \tab 3 \tab 4 \tab 2 \tab 1 \tab 6 \tab 34 \tab 15 \tab 7 \tab 2 \tab 4 \tab 21 \tab 1 \tab - \tab 11 \tab 9 \tab 653 \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1490 \cr
   1976 \tab - \tab - \tab 21 \tab 5 \tab 10 \tab 22 \tab 26 \tab - \tab 35 \tab 16 \tab 7 \tab 13 \tab 14 \tab 6 \tab 17 \tab 7 \tab 35 \tab 22 \tab 20 \tab 45 \tab 3 \tab 6 \tab 11 \tab 5 \tab 8 \tab 4 \tab 21 \tab 18 \tab 25 \tab 5 \tab - \tab 13 \tab 6 \tab 2 \tab 4 \tab 6 \tab 4 \tab 18 \tab 21 \tab - \tab 33 \tab 2 \tab 3 \tab 7 \tab 14 \tab 47 \tab 4 \tab 32 \tab 3 \tab - \tab 1 \tab 5 \tab 18 \tab 5 \tab 2 \tab 9 \tab 3 \tab 3 \tab 7 \tab 2 \tab - \tab 2 \tab - \tab 7 \tab 1 \tab 1 \tab 10 \tab 2 \tab 1 \tab 1 \tab 8 \tab - \tab 3 \tab 3 \tab 31 \tab 12 \tab 3 \tab 3 \tab 7 \tab 19 \tab 3 \tab - \tab 15 \tab 3 \tab 646 \tab 3 \tab 6 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab 1499 \cr
   1977 \tab 3 \tab 4 \tab 10 \tab 5 \tab 17 \tab 14 \tab 26 \tab 3 \tab 22 \tab 6 \tab 3 \tab 9 \tab 16 \tab 13 \tab 11 \tab 3 \tab 24 \tab 17 \tab 16 \tab 62 \tab 3 \tab 8 \tab 6 \tab 4 \tab 13 \tab 5 \tab 10 \tab 17 \tab 12 \tab 3 \tab 1 \tab 8 \tab 2 \tab 1 \tab 8 \tab 11 \tab 4 \tab 29 \tab 14 \tab 1 \tab 27 \tab 6 \tab 2 \tab 7 \tab 15 \tab 49 \tab 8 \tab 26 \tab 6 \tab 3 \tab - \tab 4 \tab 18 \tab 9 \tab 2 \tab 9 \tab 2 \tab 8 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 5 \tab 2 \tab 5 \tab 7 \tab 3 \tab - \tab 2 \tab 7 \tab 1 \tab 2 \tab 2 \tab 29 \tab 12 \tab 3 \tab 4 \tab 9 \tab 24 \tab 3 \tab - \tab 13 \tab 3 \tab 733 \tab - \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1530 \cr
   1978 \tab - \tab 4 \tab 16 \tab 7 \tab 12 \tab 20 \tab 23 \tab 4 \tab 31 \tab 7 \tab 12 \tab 6 \tab 17 \tab 5 \tab 9 \tab 5 \tab 26 \tab 19 \tab 15 \tab 39 \tab 5 \tab 8 \tab 6 \tab 6 \tab 12 \tab 6 \tab 17 \tab 12 \tab 29 \tab 8 \tab - \tab 13 \tab 10 \tab 4 \tab 7 \tab 2 \tab 8 \tab 14 \tab 18 \tab 1 \tab 43 \tab 7 \tab 6 \tab 6 \tab 13 \tab 55 \tab 3 \tab 24 \tab 4 \tab 2 \tab 1 \tab 9 \tab 21 \tab 15 \tab 8 \tab 10 \tab 6 \tab 3 \tab 6 \tab 1 \tab 2 \tab - \tab 7 \tab 5 \tab 1 \tab 5 \tab 7 \tab 3 \tab 5 \tab 7 \tab 7 \tab 5 \tab - \tab 4 \tab 40 \tab 15 \tab 6 \tab 1 \tab 8 \tab 13 \tab 2 \tab - \tab 11 \tab 2 \tab 660 \tab 2 \tab 2 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 2 \tab 4 \tab 11 \tab 3 \tab 10 \tab 19 \tab 15 \tab 1 \tab 31 \tab 7 \tab 7 \tab 5 \tab 11 \tab 7 \tab 11 \tab 4 \tab 20 \tab 15 \tab 23 \tab 49 \tab 1 \tab 6 \tab 6 \tab 1 \tab 7 \tab 1 \tab 12 \tab 19 \tab 19 \tab 15 \tab 1 \tab 14 \tab 9 \tab 4 \tab 7 \tab 4 \tab 2 \tab 19 \tab 9 \tab 4 \tab 37 \tab 5 \tab 1 \tab 5 \tab 16 \tab 45 \tab 2 \tab 19 \tab 6 \tab 5 \tab 1 \tab 6 \tab 24 \tab 14 \tab 6 \tab 11 \tab 2 \tab 5 \tab 3 \tab - \tab 1 \tab 3 \tab 2 \tab 5 \tab - \tab 5 \tab 7 \tab 6 \tab 1 \tab 3 \tab 7 \tab 3 \tab - \tab 5 \tab 29 \tab 8 \tab 6 \tab 1 \tab 9 \tab 18 \tab 2 \tab 1 \tab 12 \tab 2 \tab 686 \tab - \tab 7 \tab 5 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab - \tab 1468 \cr
   1982 \tab 2 \tab 6 \tab 23 \tab 8 \tab 15 \tab 20 \tab 31 \tab 8 \tab 35 \tab 9 \tab 5 \tab 9 \tab 13 \tab 8 \tab 11 \tab 2 \tab 24 \tab 29 \tab 33 \tab 50 \tab 1 \tab 6 \tab 4 \tab 10 \tab 15 \tab 4 \tab 15 \tab 12 \tab 26 \tab 7 \tab 1 \tab 12 \tab 12 \tab 5 \tab 7 \tab 8 \tab 5 \tab 19 \tab 17 \tab 1 \tab 46 \tab 10 \tab 2 \tab 7 \tab 24 \tab 47 \tab 5 \tab 18 \tab 3 \tab 2 \tab 2 \tab 8 \tab 13 \tab 8 \tab 3 \tab 9 \tab 5 \tab 6 \tab 5 \tab 2 \tab - \tab 3 \tab 1 \tab 4 \tab - \tab 8 \tab 5 \tab 6 \tab 2 \tab 1 \tab 9 \tab 2 \tab 3 \tab 2 \tab 34 \tab 12 \tab 2 \tab 4 \tab 3 \tab 15 \tab 1 \tab 1 \tab 12 \tab 7 \tab 962 \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 2 \tab 6 \tab 14 \tab 6 \tab 8 \tab 16 \tab 29 \tab 3 \tab 32 \tab 8 \tab 2 \tab 16 \tab 12 \tab 9 \tab 8 \tab 5 \tab 28 \tab 21 \tab 19 \tab 61 \tab 3 \tab 8 \tab 3 \tab 1 \tab 8 \tab 3 \tab 7 \tab 16 \tab 26 \tab 5 \tab 1 \tab 16 \tab 7 \tab 5 \tab 1 \tab 4 \tab 4 \tab 28 \tab 13 \tab - \tab 38 \tab 10 \tab 1 \tab 2 \tab 13 \tab 68 \tab 8 \tab 19 \tab 5 \tab 2 \tab - \tab 8 \tab 13 \tab 13 \tab 10 \tab 10 \tab 5 \tab 7 \tab 11 \tab 1 \tab 1 \tab 2 \tab 2 \tab 8 \tab 1 \tab 5 \tab 2 \tab 8 \tab 2 \tab 2 \tab 5 \tab 2 \tab 2 \tab 2 \tab 19 \tab 18 \tab 8 \tab 3 \tab 13 \tab 22 \tab 1 \tab - \tab 17 \tab 7 \tab 722 \tab - \tab 6 \tab 2 \tab 3 \tab - \tab - \tab 3 \tab 3 \tab - \tab 1599 \cr
   1984 \tab 3 \tab 6 \tab 12 \tab 5 \tab 7 \tab 17 \tab 35 \tab 3 \tab 34 \tab 5 \tab 7 \tab 7 \tab 15 \tab 8 \tab 10 \tab 2 \tab 25 \tab 17 \tab 12 \tab 43 \tab - \tab 7 \tab 2 \tab 8 \tab 8 \tab 2 \tab 10 \tab 11 \tab 30 \tab 4 \tab - \tab 13 \tab 4 \tab 3 \tab 6 \tab 4 \tab 3 \tab 16 \tab 9 \tab 1 \tab 32 \tab 3 \tab 2 \tab 14 \tab 15 \tab 42 \tab 4 \tab 20 \tab 3 \tab 3 \tab 1 \tab 3 \tab 18 \tab 16 \tab 8 \tab 13 \tab 3 \tab 4 \tab 4 \tab 1 \tab - \tab 2 \tab 3 \tab 5 \tab 2 \tab 7 \tab 12 \tab 7 \tab 1 \tab 2 \tab 10 \tab 4 \tab 4 \tab 6 \tab 15 \tab 13 \tab 6 \tab 3 \tab 9 \tab 14 \tab 1 \tab 1 \tab 14 \tab 2 \tab 695 \tab - \tab 2 \tab 8 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1473 \cr
   1985 \tab 1 \tab 2 \tab 22 \tab 9 \tab 9 \tab 9 \tab 30 \tab 1 \tab 44 \tab 8 \tab 4 \tab 4 \tab 16 \tab 3 \tab 11 \tab 6 \tab 22 \tab 18 \tab 24 \tab 49 \tab 3 \tab 5 \tab 4 \tab 3 \tab 9 \tab - \tab 8 \tab 11 \tab 24 \tab 4 \tab 1 \tab 11 \tab 8 \tab 4 \tab 2 \tab 6 \tab 4 \tab 33 \tab 10 \tab 2 \tab 29 \tab 4 \tab 2 \tab 6 \tab 19 \tab 47 \tab 6 \tab 23 \tab 5 \tab 1 \tab - \tab 5 \tab 20 \tab 16 \tab 5 \tab 11 \tab 5 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab 5 \tab - \tab 5 \tab 1 \tab 2 \tab 2 \tab - \tab 10 \tab 1 \tab 3 \tab 6 \tab 27 \tab 16 \tab 4 \tab 3 \tab 7 \tab 19 \tab 2 \tab - \tab 9 \tab 6 \tab 739 \tab 1 \tab 4 \tab 3 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 1534 \cr
   1986 \tab 1 \tab 4 \tab 13 \tab 5 \tab 9 \tab 9 \tab 34 \tab 1 \tab 34 \tab 9 \tab 7 \tab 6 \tab 6 \tab 5 \tab 6 \tab 2 \tab 15 \tab 14 \tab 15 \tab 54 \tab 4 \tab 6 \tab 7 \tab 3 \tab 12 \tab 3 \tab 9 \tab 5 \tab 20 \tab 1 \tab 1 \tab 9 \tab 4 \tab 4 \tab 9 \tab 1 \tab 4 \tab 13 \tab 19 \tab 3 \tab 23 \tab 7 \tab 2 \tab 10 \tab 15 \tab 52 \tab 8 \tab 23 \tab 8 \tab 4 \tab 2 \tab 3 \tab 17 \tab 9 \tab 7 \tab 5 \tab 5 \tab 3 \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 7 \tab - \tab 3 \tab 7 \tab 3 \tab 2 \tab 1 \tab 14 \tab 2 \tab 3 \tab 7 \tab 32 \tab 18 \tab 3 \tab 2 \tab 8 \tab 15 \tab 4 \tab - \tab 10 \tab 6 \tab 712 \tab - \tab 6 \tab 6 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1470 \cr
   1987 \tab 1 \tab 3 \tab 15 \tab 4 \tab 11 \tab 15 \tab 30 \tab 2 \tab 40 \tab 8 \tab 9 \tab 6 \tab 11 \tab 11 \tab 8 \tab 3 \tab 22 \tab 17 \tab 36 \tab 57 \tab 2 \tab 2 \tab 4 \tab 5 \tab 9 \tab 3 \tab 14 \tab 14 \tab 33 \tab 6 \tab - \tab 17 \tab 6 \tab 3 \tab 4 \tab 4 \tab 4 \tab 8 \tab 14 \tab - \tab 33 \tab 5 \tab 3 \tab 10 \tab 9 \tab 33 \tab 6 \tab 31 \tab 3 \tab 2 \tab 3 \tab 8 \tab 16 \tab 13 \tab 5 \tab 12 \tab 4 \tab 5 \tab 4 \tab 1 \tab - \tab 3 \tab 3 \tab 7 \tab - \tab 5 \tab 9 \tab 4 \tab 1 \tab 2 \tab 4 \tab 2 \tab 2 \tab 1 \tab 34 \tab 12 \tab 2 \tab 3 \tab 13 \tab 18 \tab 2 \tab - \tab 10 \tab 4 \tab 1005 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 2 \tab 4 \tab 12 \tab 3 \tab 9 \tab 9 \tab 25 \tab 1 \tab 28 \tab 1 \tab 9 \tab 10 \tab 6 \tab 7 \tab 5 \tab 1 \tab 18 \tab 18 \tab 29 \tab 43 \tab 1 \tab 9 \tab 9 \tab 3 \tab 4 \tab 1 \tab 10 \tab 12 \tab 21 \tab 3 \tab - \tab 18 \tab 6 \tab 6 \tab 2 \tab 4 \tab 3 \tab 15 \tab 15 \tab 1 \tab 26 \tab 5 \tab 2 \tab 5 \tab 19 \tab 41 \tab 7 \tab 17 \tab 10 \tab 2 \tab 2 \tab 5 \tab 9 \tab 12 \tab 7 \tab 7 \tab 2 \tab 4 \tab 11 \tab 1 \tab 1 \tab 2 \tab 2 \tab 7 \tab - \tab 5 \tab 4 \tab 7 \tab - \tab 2 \tab 7 \tab 4 \tab 3 \tab 8 \tab 20 \tab 12 \tab 3 \tab 4 \tab 7 \tab 20 \tab - \tab 1 \tab 7 \tab 9 \tab 751 \tab 1 \tab 5 \tab 7 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1481 \cr
   1989 \tab 1 \tab 3 \tab 9 \tab 1 \tab 3 \tab 17 \tab 29 \tab 1 \tab 21 \tab 12 \tab 8 \tab 8 \tab 18 \tab 13 \tab 11 \tab 6 \tab 21 \tab 19 \tab 26 \tab 45 \tab 1 \tab 6 \tab 8 \tab 2 \tab 9 \tab 5 \tab 13 \tab 12 \tab 20 \tab 5 \tab 1 \tab 19 \tab 3 \tab 7 \tab 2 \tab 3 \tab 4 \tab 17 \tab 11 \tab - \tab 24 \tab 8 \tab 2 \tab 7 \tab 15 \tab 58 \tab 4 \tab 24 \tab 4 \tab 1 \tab 1 \tab 8 \tab 14 \tab 9 \tab 3 \tab 11 \tab 2 \tab 5 \tab 3 \tab 1 \tab 1 \tab - \tab 5 \tab 14 \tab 1 \tab 6 \tab 12 \tab 5 \tab - \tab 1 \tab 7 \tab 4 \tab 2 \tab 2 \tab 19 \tab 17 \tab 4 \tab 2 \tab 5 \tab 13 \tab 3 \tab - \tab 14 \tab 5 \tab 751 \tab - \tab 8 \tab 9 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1537 \cr
   1990 \tab 3 \tab 1 \tab 9 \tab 3 \tab 7 \tab 7 \tab 19 \tab 1 \tab 19 \tab 4 \tab 5 \tab 6 \tab 10 \tab 12 \tab 8 \tab 2 \tab 16 \tab 16 \tab 25 \tab 28 \tab 1 \tab 3 \tab 6 \tab 3 \tab 6 \tab - \tab 4 \tab 9 \tab 18 \tab 7 \tab - \tab 13 \tab 5 \tab 8 \tab 1 \tab 8 \tab 6 \tab 22 \tab 25 \tab - \tab 19 \tab 4 \tab 2 \tab 3 \tab 17 \tab 44 \tab 3 \tab 20 \tab 4 \tab 2 \tab 2 \tab 4 \tab 12 \tab 13 \tab 4 \tab 8 \tab 3 \tab 5 \tab 12 \tab - \tab 2 \tab - \tab - \tab 8 \tab - \tab 6 \tab 6 \tab 3 \tab 1 \tab - \tab 8 \tab 2 \tab 1 \tab 5 \tab 23 \tab 9 \tab 3 \tab 2 \tab 17 \tab 12 \tab 3 \tab 1 \tab 16 \tab 3 \tab 695 \tab - \tab 7 \tab 7 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1372 \cr
   1991 \tab - \tab 3 \tab 13 \tab 3 \tab 12 \tab 13 \tab 29 \tab 1 \tab 26 \tab 10 \tab 5 \tab 7 \tab 5 \tab 11 \tab 7 \tab 4 \tab 22 \tab 14 \tab 19 \tab 37 \tab 1 \tab 9 \tab 8 \tab 10 \tab 14 \tab 4 \tab 10 \tab 10 \tab 23 \tab 5 \tab - \tab 17 \tab 6 \tab 6 \tab 4 \tab 5 \tab 5 \tab 17 \tab 14 \tab 1 \tab 36 \tab 6 \tab 1 \tab 7 \tab 17 \tab 42 \tab 6 \tab 9 \tab 4 \tab 2 \tab - \tab 1 \tab 17 \tab 7 \tab 1 \tab 9 \tab 6 \tab 10 \tab 2 \tab 2 \tab - \tab - \tab 3 \tab 10 \tab - \tab 4 \tab 4 \tab 6 \tab 1 \tab - \tab 8 \tab 3 \tab 4 \tab 1 \tab 34 \tab 17 \tab 3 \tab 2 \tab 11 \tab 23 \tab 2 \tab - \tab 11 \tab 5 \tab 757 \tab - \tab 6 \tab 6 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 3 \tab 1 \tab 7 \tab 4 \tab 5 \tab 11 \tab 24 \tab 4 \tab 26 \tab 8 \tab 8 \tab 5 \tab 15 \tab 9 \tab 8 \tab 2 \tab 27 \tab 14 \tab 18 \tab 46 \tab 2 \tab 4 \tab 8 \tab 3 \tab 7 \tab 3 \tab 8 \tab 5 \tab 24 \tab 8 \tab - \tab 16 \tab 8 \tab 10 \tab 6 \tab 6 \tab 3 \tab 16 \tab 15 \tab 3 \tab 36 \tab 7 \tab 1 \tab 5 \tab 19 \tab 48 \tab 4 \tab 28 \tab 5 \tab 5 \tab 3 \tab 5 \tab 11 \tab 9 \tab 9 \tab 11 \tab 4 \tab 3 \tab 6 \tab 1 \tab 1 \tab 5 \tab 2 \tab 12 \tab 2 \tab 7 \tab 6 \tab 9 \tab 3 \tab 1 \tab 13 \tab 3 \tab - \tab 6 \tab 24 \tab 23 \tab 3 \tab 3 \tab 19 \tab 15 \tab - \tab 1 \tab 10 \tab 5 \tab 792 \tab - \tab 5 \tab 12 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1606 \cr
   1994 \tab 3 \tab 7 \tab 22 \tab 10 \tab 14 \tab 24 \tab 42 \tab 3 \tab 47 \tab 11 \tab 18 \tab 11 \tab 13 \tab 15 \tab 16 \tab 6 \tab 36 \tab 27 \tab 36 \tab 82 \tab 3 \tab 15 \tab 16 \tab 8 \tab 7 \tab 7 \tab 12 \tab 16 \tab 49 \tab 11 \tab - \tab 21 \tab 12 \tab 10 \tab 9 \tab 11 \tab 3 \tab 37 \tab 22 \tab 2 \tab 59 \tab 10 \tab 2 \tab 8 \tab 32 \tab 107 \tab 9 \tab 27 \tab 11 \tab 2 \tab 8 \tab 12 \tab 27 \tab 21 \tab 5 \tab 18 \tab 8 \tab 11 \tab 16 \tab 1 \tab 4 \tab 1 \tab 6 \tab 23 \tab 2 \tab 9 \tab 9 \tab 9 \tab 1 \tab 3 \tab 23 \tab 4 \tab 5 \tab 9 \tab 43 \tab 39 \tab 10 \tab 4 \tab 32 \tab 29 \tab 10 \tab 3 \tab 33 \tab 6 \tab 1533 \tab 1 \tab 9 \tab 10 \tab 3 \tab - \tab 3 \tab 4 \tab 4 \tab - \tab 2992 \cr
   1996 \tab 3 \tab 5 \tab 20 \tab 10 \tab 13 \tab 21 \tab 34 \tab 1 \tab 56 \tab 6 \tab 12 \tab 12 \tab 11 \tab 12 \tab 16 \tab 4 \tab 25 \tab 28 \tab 49 \tab 75 \tab 2 \tab 19 \tab 12 \tab 1 \tab 13 \tab 9 \tab 13 \tab 21 \tab 47 \tab 10 \tab - \tab 26 \tab 9 \tab 12 \tab 3 \tab 8 \tab 4 \tab 22 \tab 24 \tab 2 \tab 47 \tab 7 \tab 2 \tab 10 \tab 25 \tab 67 \tab 13 \tab 13 \tab 6 \tab 4 \tab 6 \tab 10 \tab 24 \tab 17 \tab 6 \tab 17 \tab 8 \tab 10 \tab 13 \tab 1 \tab - \tab 4 \tab 8 \tab 8 \tab 2 \tab 11 \tab 4 \tab 10 \tab 3 \tab 6 \tab 18 \tab 5 \tab 11 \tab 14 \tab 51 \tab 33 \tab 8 \tab 7 \tab 26 \tab 32 \tab 4 \tab 5 \tab 27 \tab 15 \tab 1580 \tab 2 \tab 9 \tab 16 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 2904 \cr
   1998 \tab 2 \tab 6 \tab 19 \tab 2 \tab 9 \tab 17 \tab 36 \tab - \tab 45 \tab 4 \tab 14 \tab 12 \tab 6 \tab 10 \tab 12 \tab 6 \tab 23 \tab 30 \tab 39 \tab 74 \tab 4 \tab 6 \tab 11 \tab 4 \tab 12 \tab 9 \tab 25 \tab 11 \tab 37 \tab 10 \tab 1 \tab 32 \tab 11 \tab 12 \tab 4 \tab 7 \tab 8 \tab 21 \tab 16 \tab 1 \tab 53 \tab 9 \tab 5 \tab 11 \tab 24 \tab 73 \tab 12 \tab 22 \tab 14 \tab 3 \tab 5 \tab 6 \tab 25 \tab 28 \tab 12 \tab 25 \tab 5 \tab 15 \tab 11 \tab 1 \tab 2 \tab 4 \tab 9 \tab 16 \tab 6 \tab 13 \tab 2 \tab 23 \tab 5 \tab 3 \tab 13 \tab 4 \tab 3 \tab 5 \tab 45 \tab 24 \tab 7 \tab 6 \tab 24 \tab 27 \tab 5 \tab 2 \tab 33 \tab 11 \tab 1549 \tab 1 \tab 10 \tab 9 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab 2832 \cr
   2000 \tab 1 \tab 6 \tab 17 \tab 4 \tab 10 \tab 13 \tab 39 \tab 5 \tab 34 \tab 11 \tab 20 \tab 11 \tab 5 \tab 14 \tab 21 \tab 7 \tab 37 \tab 23 \tab 24 \tab 67 \tab 5 \tab 11 \tab 5 \tab 4 \tab 12 \tab 4 \tab 12 \tab 19 \tab 39 \tab 6 \tab - \tab 23 \tab 7 \tab 10 \tab 6 \tab 7 \tab 7 \tab 28 \tab 19 \tab 3 \tab 41 \tab 4 \tab 2 \tab 7 \tab 25 \tab 47 \tab 8 \tab 17 \tab 7 \tab 6 \tab 3 \tab 10 \tab 31 \tab 19 \tab 20 \tab 18 \tab 5 \tab 16 \tab 11 \tab 1 \tab 1 \tab 2 \tab 9 \tab 11 \tab 2 \tab 8 \tab 8 \tab 9 \tab 2 \tab 3 \tab 12 \tab 4 \tab 8 \tab 10 \tab 54 \tab 15 \tab 9 \tab 5 \tab 25 \tab 19 \tab 4 \tab 2 \tab 18 \tab 9 \tab 1648 \tab 1 \tab 7 \tab 10 \tab 2 \tab - \tab - \tab 4 \tab 2 \tab - \tab 2817 \cr
   2002 \tab 1 \tab 5 \tab 17 \tab 3 \tab 15 \tab 11 \tab 41 \tab 3 \tab 37 \tab 4 \tab 19 \tab 5 \tab 8 \tab 16 \tab 13 \tab 5 \tab 25 \tab 23 \tab 45 \tab 58 \tab 7 \tab 11 \tab 6 \tab 5 \tab 8 \tab 6 \tab 10 \tab 16 \tab 33 \tab 8 \tab 2 \tab 28 \tab 8 \tab 13 \tab 3 \tab 8 \tab 5 \tab 23 \tab 19 \tab 5 \tab 50 \tab 12 \tab 1 \tab 10 \tab 33 \tab 67 \tab 9 \tab 16 \tab 16 \tab 6 \tab 2 \tab 3 \tab 23 \tab 21 \tab 17 \tab 15 \tab 3 \tab 12 \tab 10 \tab 1 \tab 6 \tab 5 \tab 5 \tab 13 \tab 2 \tab 8 \tab 8 \tab 15 \tab 3 \tab 4 \tab 20 \tab 3 \tab 5 \tab 9 \tab 43 \tab 15 \tab 8 \tab 8 \tab 16 \tab 30 \tab 5 \tab 1 \tab 22 \tab 14 \tab 1565 \tab - \tab 8 \tab 12 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2765 \cr
   2004 \tab 3 \tab 6 \tab 18 \tab 7 \tab 16 \tab 17 \tab 25 \tab 1 \tab 38 \tab 10 \tab 17 \tab 6 \tab 5 \tab 15 \tab 16 \tab 5 \tab 26 \tab 33 \tab 42 \tab 87 \tab 2 \tab 12 \tab 10 \tab 6 \tab 9 \tab 7 \tab 7 \tab 18 \tab 48 \tab 7 \tab 1 \tab 34 \tab 16 \tab 14 \tab 11 \tab 11 \tab 4 \tab 14 \tab 27 \tab 1 \tab 42 \tab 13 \tab 3 \tab 7 \tab 33 \tab 69 \tab 15 \tab 26 \tab 12 \tab 6 \tab 8 \tab 8 \tab 23 \tab 22 \tab 9 \tab 26 \tab 8 \tab 15 \tab 20 \tab - \tab 2 \tab 1 \tab 2 \tab 18 \tab 1 \tab 18 \tab 6 \tab 14 \tab 3 \tab 2 \tab 19 \tab 3 \tab 11 \tab 8 \tab 67 \tab 25 \tab 9 \tab 3 \tab 16 \tab 29 \tab 6 \tab 3 \tab 29 \tab 16 \tab 1448 \tab 1 \tab 8 \tab 16 \tab 3 \tab 1 \tab - \tab 2 \tab 5 \tab - \tab 2812 \cr
   2006 \tab 2 \tab 6 \tab 27 \tab 11 \tab 23 \tab 28 \tab 44 \tab 6 \tab 48 \tab 7 \tab 32 \tab 19 \tab 16 \tab 35 \tab 34 \tab 16 \tab 53 \tab 40 \tab 66 \tab 123 \tab 5 \tab 7 \tab 15 \tab 9 \tab 13 \tab 7 \tab 20 \tab 17 \tab 54 \tab 8 \tab - \tab 42 \tab 22 \tab 23 \tab 12 \tab 4 \tab 6 \tab 24 \tab 43 \tab 7 \tab 95 \tab 13 \tab 2 \tab 18 \tab 28 \tab 93 \tab 17 \tab 26 \tab 9 \tab 5 \tab 6 \tab 11 \tab 35 \tab 24 \tab 15 \tab 36 \tab 10 \tab 22 \tab 15 \tab 4 \tab 1 \tab 7 \tab 9 \tab 15 \tab 5 \tab 23 \tab 13 \tab 33 \tab 10 \tab 5 \tab 21 \tab 9 \tab 19 \tab 15 \tab 92 \tab 48 \tab 6 \tab 5 \tab 39 \tab 39 \tab 7 \tab 5 \tab 47 \tab 21 \tab 2508 \tab 1 \tab 16 \tab 8 \tab 6 \tab 3 \tab 4 \tab 5 \tab 7 \tab - \tab 4510 \cr
   2008 \tab 1 \tab 5 \tab 10 \tab 1 \tab 12 \tab 14 \tab 27 \tab 1 \tab 27 \tab 6 \tab 14 \tab 5 \tab 5 \tab 11 \tab 9 \tab 2 \tab 21 \tab 9 \tab 28 \tab 59 \tab 1 \tab 8 \tab 13 \tab 2 \tab 10 \tab 4 \tab 8 \tab 6 \tab 23 \tab 4 \tab - \tab 20 \tab 7 \tab 10 \tab 4 \tab 5 \tab 4 \tab 13 \tab 14 \tab 1 \tab 40 \tab 8 \tab 1 \tab 8 \tab 16 \tab 52 \tab 11 \tab 19 \tab 5 \tab 5 \tab 1 \tab 2 \tab 13 \tab 14 \tab 9 \tab 21 \tab 4 \tab 10 \tab 7 \tab - \tab - \tab 4 \tab 9 \tab 9 \tab 2 \tab 10 \tab 15 \tab 15 \tab 3 \tab 2 \tab 14 \tab 4 \tab 14 \tab 7 \tab 26 \tab 19 \tab 3 \tab 4 \tab 16 \tab 26 \tab 3 \tab 3 \tab 23 \tab 3 \tab 1095 \tab - \tab 8 \tab 9 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab - \tab 2023 \cr
   2010 \tab 1 \tab 4 \tab 12 \tab 6 \tab 7 \tab 18 \tab 23 \tab 2 \tab 25 \tab 8 \tab 12 \tab 7 \tab 5 \tab 11 \tab 9 \tab 5 \tab 18 \tab 9 \tab 24 \tab 45 \tab - \tab 7 \tab 13 \tab 2 \tab 8 \tab 6 \tab 12 \tab 11 \tab 16 \tab 2 \tab - \tab 15 \tab 9 \tab 8 \tab 4 \tab 6 \tab 1 \tab 10 \tab 16 \tab 1 \tab 30 \tab 5 \tab 5 \tab 8 \tab 22 \tab 43 \tab 4 \tab 12 \tab 8 \tab 3 \tab 1 \tab 6 \tab 14 \tab 10 \tab 5 \tab 23 \tab 8 \tab 10 \tab 10 \tab 1 \tab - \tab 4 \tab 3 \tab 7 \tab 4 \tab 12 \tab 5 \tab 9 \tab 3 \tab 1 \tab 16 \tab 2 \tab 5 \tab 5 \tab 29 \tab 15 \tab 3 \tab 3 \tab 15 \tab 17 \tab 3 \tab 3 \tab 11 \tab 12 \tab 1205 \tab 3 \tab 8 \tab 8 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 2044 \cr
   2012 \tab - \tab 1 \tab 5 \tab 2 \tab 11 \tab 17 \tab 14 \tab 1 \tab 18 \tab 5 \tab 9 \tab 5 \tab 4 \tab 14 \tab 11 \tab 3 \tab 20 \tab 30 \tab 19 \tab 42 \tab 1 \tab 2 \tab 14 \tab 5 \tab 10 \tab 3 \tab 5 \tab 8 \tab 17 \tab 8 \tab - \tab 28 \tab 4 \tab 5 \tab 6 \tab 3 \tab 2 \tab 12 \tab 9 \tab - \tab 29 \tab 3 \tab 1 \tab 7 \tab 10 \tab 40 \tab 5 \tab 9 \tab 4 \tab 3 \tab 1 \tab 6 \tab 6 \tab 10 \tab 11 \tab 18 \tab 5 \tab 7 \tab 9 \tab - \tab 1 \tab 4 \tab 5 \tab 10 \tab 4 \tab 15 \tab 1 \tab 11 \tab 4 \tab 4 \tab 13 \tab 1 \tab 11 \tab 6 \tab 43 \tab 15 \tab 3 \tab 2 \tab 16 \tab 19 \tab 3 \tab 2 \tab 25 \tab 29 \tab 1149 \tab - \tab 8 \tab 10 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 6 \tab 8 \tab 7 \tab 11 \tab 21 \tab 25 \tab 1 \tab 31 \tab 5 \tab 13 \tab 7 \tab 7 \tab 17 \tab 14 \tab 7 \tab 23 \tab 22 \tab 28 \tab 55 \tab - \tab 12 \tab 17 \tab 5 \tab 10 \tab 1 \tab 9 \tab 6 \tab 25 \tab 4 \tab 2 \tab 21 \tab 11 \tab 16 \tab 9 \tab 5 \tab 2 \tab 15 \tab 19 \tab - \tab 38 \tab 14 \tab 3 \tab 8 \tab 20 \tab 51 \tab 12 \tab 14 \tab 7 \tab 2 \tab 3 \tab 7 \tab 12 \tab 8 \tab 8 \tab 23 \tab 2 \tab 20 \tab 9 \tab 1 \tab - \tab 2 \tab 8 \tab 11 \tab 7 \tab 15 \tab 18 \tab 20 \tab 3 \tab 1 \tab 16 \tab 4 \tab 12 \tab 11 \tab 33 \tab 22 \tab 7 \tab 4 \tab 15 \tab 29 \tab 4 \tab 1 \tab 29 \tab 14 \tab 1455 \tab 4 \tab 12 \tab 11 \tab 3 \tab 2 \tab - \tab 2 \tab 4 \tab - \tab 2538 \cr
   2016 \tab 1 \tab 6 \tab 12 \tab 2 \tab 9 \tab 21 \tab 39 \tab 3 \tab 30 \tab 7 \tab 15 \tab 7 \tab 2 \tab 14 \tab 20 \tab 5 \tab 25 \tab 21 \tab 29 \tab 52 \tab 2 \tab 12 \tab 11 \tab 2 \tab 9 \tab 4 \tab 13 \tab 10 \tab 30 \tab 7 \tab 2 \tab 33 \tab 11 \tab 17 \tab 8 \tab 6 \tab 5 \tab 13 \tab 22 \tab 1 \tab 48 \tab 6 \tab 4 \tab 6 \tab 21 \tab 42 \tab 15 \tab 12 \tab 8 \tab 5 \tab 4 \tab 5 \tab 15 \tab 10 \tab 10 \tab 20 \tab 4 \tab 16 \tab 13 \tab 3 \tab 3 \tab 4 \tab 10 \tab 8 \tab 3 \tab 13 \tab 8 \tab 8 \tab 9 \tab 1 \tab 14 \tab 4 \tab 17 \tab 4 \tab 48 \tab 30 \tab 11 \tab 7 \tab 16 \tab 27 \tab 3 \tab 3 \tab 34 \tab 6 \tab 1723 \tab 1 \tab 15 \tab 17 \tab 4 \tab 1 \tab 1 \tab 6 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 2 \tab 12 \tab 6 \tab 9 \tab 10 \tab 26 \tab 24 \tab 6 \tab 29 \tab 3 \tab 8 \tab 3 \tab 5 \tab 20 \tab 25 \tab 4 \tab 30 \tab 21 \tab 38 \tab 62 \tab - \tab 19 \tab 15 \tab 2 \tab 13 \tab 7 \tab 12 \tab 15 \tab 28 \tab 1 \tab - \tab 27 \tab 11 \tab 9 \tab 4 \tab 7 \tab 8 \tab 8 \tab 24 \tab 4 \tab 46 \tab 8 \tab 3 \tab 11 \tab 28 \tab 41 \tab 14 \tab 20 \tab 13 \tab 4 \tab 3 \tab 8 \tab 8 \tab 22 \tab 10 \tab 18 \tab 7 \tab 15 \tab 14 \tab 1 \tab 4 \tab 3 \tab 7 \tab 12 \tab 5 \tab 12 \tab 4 \tab 15 \tab 9 \tab 2 \tab 13 \tab 6 \tab 14 \tab 8 \tab 36 \tab 32 \tab 6 \tab 3 \tab 15 \tab 33 \tab 6 \tab 2 \tab 28 \tab 10 \tab 1179 \tab - \tab 11 \tab 10 \tab 5 \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 51 \tab 141 \tab 467 \tab 180 \tab 370 \tab 520 \tab 962 \tab 71 \tab 1068 \tab 253 \tab 326 \tab 283 \tab 342 \tab 376 \tab 414 \tab 150 \tab 819 \tab 653 \tab 866 \tab 1805 \tab 79 \tab 266 \tab 285 \tab 135 \tab 319 \tab 127 \tab 370 \tab 419 \tab 865 \tab 195 \tab 17 \tab 592 \tab 255 \tab 250 \tab 171 \tab 195 \tab 134 \tab 617 \tab 579 \tab 51 \tab 1233 \tab 221 \tab 76 \tab 245 \tab 624 \tab 1754 \tab 233 \tab 663 \tab 218 \tab 96 \tab 73 \tab 207 \tab 570 \tab 437 \tab 236 \tab 452 \tab 150 \tab 271 \tab 265 \tab 32 \tab 35 \tab 71 \tab 132 \tab 290 \tab 58 \tab 252 \tab 226 \tab 293 \tab 82 \tab 74 \tab 359 \tab 98 \tab 173 \tab 197 \tab 1160 \tab 610 \tab 182 \tab 110 \tab 442 \tab 695 \tab 101 \tab 42 \tab 586 \tab 260 \tab 37061 \tab 26 \tab 212 \tab 240 \tab 52 \tab 16 \tab 23 \tab 55 \tab 51 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
