% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_40.R
\name{sprel}
\alias{sprel}
\title{Spouse's religious preference}
\source{
General Social Survey https://gss.norc.org
}
\description{
sprel
}
\details{
Question 115. In what religion were you raised?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab catholic \tab jewish \tab no answer \tab none \tab other \tab protestant \tab don't know \tab buddhism \tab christian \tab hinduism \tab inter-nondenominational \tab muslim/islam \tab native american \tab other eastern religions \tab orthodox-christian \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 428 \tab 264 \tab 26 \tab 8 \tab 56 \tab 14 \tab 708 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 419 \tab 261 \tab 43 \tab 3 \tab 46 \tab 7 \tab 705 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 488 \tab 247 \tab 17 \tab - \tab 50 \tab 10 \tab 676 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 525 \tab 253 \tab 16 \tab 2 \tab 61 \tab 8 \tab 633 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 555 \tab 257 \tab 20 \tab 3 \tab 47 \tab 9 \tab 639 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 572 \tab 258 \tab 17 \tab 4 \tab 55 \tab 11 \tab 614 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 579 \tab 225 \tab 20 \tab 3 \tab 46 \tab 18 \tab 568 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 867 \tab 231 \tab 21 \tab 8 \tab 60 \tab 12 \tab 659 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 634 \tab 264 \tab 24 \tab 5 \tab 58 \tab 13 \tab 601 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 644 \tab 210 \tab 16 \tab 11 \tab 40 \tab 16 \tab 536 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 663 \tab 228 \tab 25 \tab 5 \tab 43 \tab 11 \tab 559 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 644 \tab 224 \tab 20 \tab 4 \tab 45 \tab 17 \tab 513 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 903 \tab 188 \tab 19 \tab 7 \tab 54 \tab 21 \tab 627 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 693 \tab 176 \tab 15 \tab - \tab 60 \tab 22 \tab 515 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 690 \tab 227 \tab 10 \tab 7 \tab 59 \tab 27 \tab 516 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 645 \tab 191 \tab 17 \tab 4 \tab 52 \tab 20 \tab 442 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 713 \tab 222 \tab 15 \tab 2 \tab 55 \tab 11 \tab 499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 747 \tab 198 \tab 23 \tab 8 \tab 70 \tab 23 \tab 536 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2213 \tab 188 \tab 9 \tab 18 \tab 67 \tab 28 \tab 468 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2133 \tab 174 \tab 11 \tab 1 \tab 91 \tab 1 \tab 356 \tab - \tab 3 \tab 24 \tab 3 \tab 4 \tab 5 \tab 1 \tab 5 \tab - \tab - \tab 2812 \cr
   2006 \tab 3095 \tab 404 \tab 32 \tab 1 \tab 184 \tab 6 \tab 718 \tab - \tab 8 \tab 26 \tab 7 \tab - \tab 14 \tab 3 \tab 8 \tab 4 \tab - \tab 4510 \cr
   2008 \tab 1046 \tab 227 \tab 22 \tab 9 \tab 124 \tab 7 \tab 522 \tab 3 \tab 5 \tab 31 \tab 5 \tab - \tab 6 \tab - \tab 9 \tab 7 \tab - \tab 2023 \cr
   2010 \tab 1152 \tab 224 \tab 23 \tab 1 \tab 128 \tab 2 \tab 444 \tab - \tab 7 \tab 26 \tab 2 \tab 7 \tab 11 \tab - \tab 9 \tab 8 \tab - \tab 2044 \cr
   2012 \tab 1074 \tab 217 \tab 18 \tab 2 \tab 125 \tab 3 \tab 451 \tab 10 \tab 2 \tab 45 \tab 7 \tab - \tab 11 \tab 3 \tab 3 \tab 3 \tab - \tab 1974 \cr
   2014 \tab 1376 \tab 302 \tab 17 \tab 12 \tab 198 \tab 6 \tab 526 \tab 7 \tab 4 \tab 63 \tab 12 \tab 1 \tab 9 \tab - \tab - \tab 5 \tab - \tab 2538 \cr
   2016 \tab 1654 \tab 316 \tab 25 \tab 7 \tab 213 \tab 11 \tab 552 \tab 18 \tab 5 \tab 29 \tab 11 \tab 3 \tab 10 \tab 3 \tab 5 \tab 5 \tab - \tab 2867 \cr
   2018 \tab 1348 \tab 233 \tab 20 \tab 3 \tab 182 \tab 8 \tab 477 \tab 5 \tab 8 \tab 15 \tab 5 \tab 1 \tab 22 \tab 1 \tab 17 \tab 3 \tab - \tab 2348 \cr
   2021 \tab 2053 \tab 466 \tab 27 \tab 3 \tab 487 \tab 44 \tab 801 \tab 24 \tab 20 \tab 47 \tab 25 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 30 \tab 4032 \cr
   2022 \tab 2073 \tab 361 \tab 29 \tab 22 \tab 326 \tab 17 \tab 604 \tab 16 \tab 11 \tab 22 \tab 16 \tab - \tab 19 \tab - \tab 3 \tab 15 \tab 10 \tab 3544 \cr
   Total \tab 43557 \tab 7236 \tab 597 \tab 163 \tab 3082 \tab 403 \tab 16465 \tab 105 \tab 73 \tab 328 \tab 93 \tab 19 \tab 107 \tab 12 \tab 60 \tab 50 \tab 40 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{6} buddhism
\item \code{7} hinduism
\item \code{8} other eastern religions
\item \code{9} muslim/islam
\item \code{10} orthodox-christian
\item \code{11} christian
\item \code{12} native american
\item \code{13} inter-nondenominational
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
