% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_41.R
\name{sprel16}
\alias{sprel16}
\title{Religion in which spouse raised}
\source{
General Social Survey https://gss.norc.org
}
\description{
sprel16
}
\details{
Question 117. In what religion was your (husband/wife) raised?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab catholic \tab jewish \tab no answer \tab none \tab other \tab protestant \tab don't know \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 428 \tab 268 \tab 24 \tab 9 \tab 34 \tab 9 \tab 732 \tab - \tab 1504 \cr
   1974 \tab 419 \tab 270 \tab 41 \tab 5 \tab 29 \tab 7 \tab 708 \tab 5 \tab 1484 \cr
   1975 \tab 488 \tab 254 \tab 17 \tab - \tab 27 \tab 8 \tab 686 \tab 10 \tab 1490 \cr
   1976 \tab 525 \tab 267 \tab 19 \tab 2 \tab 24 \tab 7 \tab 648 \tab 7 \tab 1499 \cr
   1977 \tab 555 \tab 264 \tab 20 \tab 2 \tab 21 \tab 7 \tab 651 \tab 10 \tab 1530 \cr
   1978 \tab 572 \tab 248 \tab 17 \tab 5 \tab 27 \tab 11 \tab 644 \tab 8 \tab 1532 \cr
   1980 \tab 579 \tab 233 \tab 18 \tab 4 \tab 26 \tab 14 \tab 575 \tab 19 \tab 1468 \cr
   1982 \tab 867 \tab 250 \tab 23 \tab 11 \tab 29 \tab 14 \tab 665 \tab 1 \tab 1860 \cr
   1983 \tab 634 \tab 274 \tab 24 \tab 8 \tab 45 \tab 10 \tab 600 \tab 4 \tab 1599 \cr
   1984 \tab 644 \tab 228 \tab 16 \tab 14 \tab 36 \tab 14 \tab 521 \tab - \tab 1473 \cr
   1985 \tab 663 \tab 241 \tab 26 \tab 8 \tab 27 \tab 10 \tab 554 \tab 5 \tab 1534 \cr
   1986 \tab 644 \tab 243 \tab 24 \tab 8 \tab 31 \tab 17 \tab 494 \tab 9 \tab 1470 \cr
   1987 \tab 903 \tab 223 \tab 20 \tab 11 \tab 26 \tab 16 \tab 618 \tab 2 \tab 1819 \cr
   1988 \tab 693 \tab 183 \tab 15 \tab 1 \tab 45 \tab 22 \tab 521 \tab 1 \tab 1481 \cr
   1989 \tab 690 \tab 256 \tab 11 \tab 7 \tab 33 \tab 22 \tab 517 \tab 1 \tab 1537 \cr
   1990 \tab 645 \tab 216 \tab 15 \tab 5 \tab 36 \tab 14 \tab 439 \tab 2 \tab 1372 \cr
   1991 \tab 713 \tab 240 \tab 16 \tab 5 \tab 34 \tab 12 \tab 497 \tab - \tab 1517 \cr
   1993 \tab 747 \tab 227 \tab 19 \tab 10 \tab 41 \tab 20 \tab 539 \tab 3 \tab 1606 \cr
   1994 \tab 2213 \tab 218 \tab 7 \tab 23 \tab 39 \tab 41 \tab 444 \tab 7 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1952 \tab 4 \tab 12 \tab - \tab - \tab 1 \tab 5 \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 55223 \tab 4607 \tab 384 \tab 138 \tab 610 \tab 276 \tab 11058 \tab 94 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
