% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_597.R
\name{spsei}
\alias{spsei}
\title{spsei}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's spouse's socioeconomic index (1980)
}
\details{
Question 1650. Socioeconomic index of respondent's spouse.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 17.1 \tab 17.6 \tab 18.5 \tab 20.1 \tab 21.2 \tab 21.6 \tab 22.5 \tab 22.6 \tab 22.7 \tab 22.9 \tab 24.2 \tab 25.5 \tab 25.8 \tab 26.4 \tab 26.5 \tab 26.7 \tab 26.9 \tab 27.2 \tab 27.5 \tab 27.7 \tab 28 \tab 28.2 \tab 28.4 \tab 28.6 \tab 28.8 \tab 29 \tab 29.2 \tab 29.3 \tab 29.4 \tab 29.5 \tab 29.6 \tab 29.8 \tab 30.1 \tab 30.4 \tab 30.7 \tab 30.8 \tab 31.2 \tab 31.3 \tab 31.4 \tab 31.5 \tab 32 \tab 32.1 \tab 32.3 \tab 32.4 \tab 32.7 \tab 33.1 \tab 33.3 \tab 33.6 \tab 33.9 \tab 34.2 \tab 34.3 \tab 34.4 \tab 34.6 \tab 34.8 \tab 35.1 \tab 35.4 \tab 35.7 \tab 36.4 \tab 36.5 \tab 36.7 \tab 36.9 \tab 37 \tab 37.2 \tab 37.3 \tab 37.4 \tab 37.5 \tab 37.6 \tab 37.7 \tab 37.8 \tab 37.9 \tab 38 \tab 38.1 \tab 38.2 \tab 38.4 \tab 38.5 \tab 38.7 \tab 38.8 \tab 38.9 \tab 39 \tab 39.4 \tab 39.5 \tab 39.6 \tab 40 \tab 40.1 \tab 41.6 \tab 42.2 \tab 43.2 \tab 43.4 \tab 43.7 \tab 44.7 \tab 44.8 \tab 45.4 \tab 45.8 \tab 45.9 \tab 46.7 \tab 47 \tab 48.5 \tab 48.6 \tab 48.8 \tab 48.9 \tab 49.3 \tab 49.6 \tab 50.5 \tab 50.7 \tab 51 \tab 51.9 \tab 52.3 \tab 52.4 \tab 52.5 \tab 53.1 \tab 53.3 \tab 53.4 \tab 53.7 \tab 53.9 \tab 54.2 \tab 54.7 \tab 55.1 \tab 56.6 \tab 57.1 \tab 58.3 \tab 59 \tab 60.5 \tab 60.6 \tab 61.4 \tab 61.9 \tab 62 \tab 62.1 \tab 62.2 \tab 62.5 \tab 63.2 \tab 63.5 \tab 64.1 \tab 64.2 \tab 64.8 \tab 65.8 \tab 66.1 \tab 67 \tab 68.1 \tab 69.2 \tab 69.7 \tab 70 \tab 71.2 \tab 71.6 \tab 72 \tab 72.5 \tab 73.2 \tab 73.3 \tab 73.6 \tab 73.9 \tab 74 \tab 74.4 \tab 75.2 \tab 75.3 \tab 75.9 \tab 76.3 \tab 76.4 \tab 76.5 \tab 78 \tab 78.5 \tab 78.8 \tab 80.3 \tab 80.9 \tab 81 \tab 83.7 \tab 84 \tab 84.2 \tab 85.1 \tab 86 \tab 86.5 \tab 86.6 \tab 86.8 \tab 86.9 \tab 87.9 \tab 88.8 \tab 91.3 \tab 92.3 \tab 92.8 \tab 93.1 \tab 97.2 \tab 24.6 \tab 24.8 \tab 25.2 \tab 25.3 \tab 25.4 \tab 25.9 \tab 26.6 \tab 27.9 \tab 29.1 \tab 29.7 \tab 30 \tab 31.7 \tab 32.8 \tab 33.2 \tab 34 \tab 35.2 \tab 36.2 \tab 39.8 \tab 41.1 \tab 41.3 \tab 42.4 \tab 43.6 \tab 44.2 \tab 44.3 \tab 44.4 \tab 44.5 \tab 45.2 \tab 46.3 \tab 47.8 \tab 47.9 \tab 50.2 \tab 54.5 \tab 60.3 \tab 64.4 \tab 65.2 \tab 68.8 \tab 73.7 \tab 73.8 \tab 74.9 \tab 79.8 \tab 79.9 \tab 80 \tab 80.1 \tab 81.9 \tab 82.7 \tab 85.5 \tab 87 \tab 87.4 \tab 88.2 \tab 96 \tab 19.4 \tab 23.3 \tab 31.1 \tab 32.9 \tab 33 \tab 36.1 \tab 39.9 \tab 40.5 \tab 42.3 \tab 49.4 \tab 49.9 \tab 53.2 \tab 56.5 \tab 58.1 \tab 59.9 \tab 61.1 \tab 70.5 \tab 71.5 \tab 83.2 \tab 83.9 \tab 84.5 \tab 84.7 \tab 85.7 \tab 90.1 \tab 93.3 \tab 21.1 \tab 30.2 \tab 33.8 \tab 36 \tab 42.9 \tab 45 \tab 47.6 \tab 50.3 \tab 51.8 \tab 53.5 \tab 54.9 \tab 55.3 \tab 59.3 \tab 64 \tab 81.3 \tab 87.5 \tab 21.3 \tab 28.9 \tab 31.9 \tab 33.5 \tab 40.4 \tab 40.9 \tab 41.2 \tab 42.7 \tab 45.3 \tab 46.8 \tab 71.3 \tab 93.7 \tab 27.3 \tab 30.6 \tab 40.2 \tab 42.5 \tab 43.3 \tab 45.7 \tab 48.4 \tab 58.8 \tab 59.1 \tab 60.4 \tab 74.1 \tab 87.2 \tab 88.3 \tab 90.3 \tab 78.9 \tab 90.8 \tab 91.7 \tab 92.9 \tab 42 \tab 66.3 \tab 74.6 \tab 77.8 \tab 89.2 \tab 91.2 \tab 27.4 \tab 34.7 \tab 49.7 \tab 86.7 \tab 88.6 \tab 37.1 \tab 39.7 \tab 46 \tab 90 \tab 42.1 \tab 43.5 \tab 46.1 \tab 85 \tab 30.5 \tab 35.6 \tab 38.3 \tab 43.1 \tab 48.7 \tab 78.7 \tab 79.3 \tab 85.8 \tab 87.3 \tab 88.9 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 751 \tab 6 \tab 2 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 12 \tab 3 \tab 8 \tab 1 \tab 3 \tab 4 \tab 7 \tab 1 \tab 1 \tab 11 \tab 5 \tab 1 \tab 4 \tab 14 \tab 9 \tab 4 \tab 7 \tab 1 \tab 5 \tab 2 \tab 1 \tab 4 \tab 5 \tab 1 \tab 3 \tab 5 \tab 9 \tab 2 \tab 2 \tab 9 \tab 7 \tab 2 \tab 10 \tab 4 \tab 3 \tab 11 \tab 11 \tab 1 \tab 3 \tab 2 \tab 4 \tab 11 \tab 6 \tab 1 \tab 2 \tab 14 \tab 1 \tab 5 \tab 1 \tab 1 \tab 14 \tab 2 \tab 1 \tab 2 \tab 7 \tab 1 \tab 2 \tab 10 \tab 5 \tab 4 \tab 41 \tab 4 \tab 3 \tab 2 \tab 13 \tab 5 \tab 4 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 2 \tab 3 \tab 7 \tab 1 \tab 2 \tab 1 \tab 4 \tab 5 \tab 1 \tab 7 \tab 2 \tab 2 \tab 4 \tab 1 \tab 8 \tab 1 \tab 21 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 3 \tab 4 \tab 4 \tab 2 \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab 2 \tab 3 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 5 \tab 9 \tab 48 \tab 1 \tab 10 \tab 3 \tab 6 \tab 3 \tab 3 \tab 3 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 6 \tab 7 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab 15 \tab 1 \tab 11 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 760 \tab 6 \tab - \tab - \tab 1 \tab 4 \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 8 \tab 1 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 3 \tab 10 \tab 1 \tab 1 \tab 12 \tab 9 \tab - \tab - \tab 12 \tab 7 \tab 6 \tab 1 \tab 2 \tab 6 \tab 1 \tab - \tab 2 \tab 5 \tab - \tab 2 \tab 6 \tab 10 \tab 5 \tab - \tab 16 \tab 8 \tab 1 \tab 21 \tab 6 \tab 8 \tab 2 \tab 9 \tab 2 \tab 1 \tab 2 \tab 1 \tab 4 \tab 3 \tab 1 \tab 2 \tab 17 \tab 1 \tab - \tab 1 \tab 1 \tab 23 \tab 1 \tab - \tab 1 \tab 9 \tab 2 \tab 1 \tab 8 \tab 4 \tab 4 \tab 56 \tab - \tab - \tab 1 \tab 12 \tab 3 \tab 4 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 8 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 4 \tab 1 \tab 14 \tab 2 \tab 2 \tab 3 \tab 1 \tab 4 \tab 1 \tab 34 \tab 2 \tab 4 \tab 1 \tab 1 \tab 6 \tab - \tab 3 \tab - \tab 5 \tab 5 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 4 \tab - \tab - \tab 2 \tab 9 \tab 9 \tab 46 \tab 5 \tab - \tab 6 \tab 6 \tab 1 \tab 3 \tab - \tab 4 \tab 4 \tab - \tab - \tab 2 \tab - \tab 1 \tab 13 \tab 4 \tab 7 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab 3 \tab 4 \tab - \tab 6 \tab 10 \tab - \tab 5 \tab 3 \tab 2 \tab 8 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab 3 \tab - \tab - \tab 7 \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 7 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 698 \tab 7 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 9 \tab - \tab 8 \tab 2 \tab - \tab 9 \tab 3 \tab 1 \tab - \tab 6 \tab 12 \tab - \tab - \tab 11 \tab 10 \tab 9 \tab 1 \tab - \tab 4 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 7 \tab 8 \tab - \tab 6 \tab 5 \tab 1 \tab 15 \tab 1 \tab 4 \tab 5 \tab 17 \tab 1 \tab - \tab 6 \tab 3 \tab 2 \tab 5 \tab - \tab - \tab 14 \tab - \tab 4 \tab 1 \tab 2 \tab 17 \tab 1 \tab 1 \tab 4 \tab 7 \tab 1 \tab 2 \tab 15 \tab 3 \tab 4 \tab 31 \tab 1 \tab 1 \tab - \tab 9 \tab 3 \tab 5 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab 3 \tab 9 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab 2 \tab 16 \tab - \tab 5 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 3 \tab 1 \tab 4 \tab 4 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 2 \tab 4 \tab 1 \tab 6 \tab 3 \tab 40 \tab 11 \tab 8 \tab 3 \tab 6 \tab 4 \tab - \tab - \tab 3 \tab 7 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 7 \tab 3 \tab - \tab 5 \tab - \tab 3 \tab 6 \tab - \tab - \tab 3 \tab 8 \tab - \tab 1 \tab 16 \tab 2 \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 4 \tab 2 \tab 4 \tab 1 \tab - \tab - \tab 4 \tab - \tab 7 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 765 \tab 5 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 5 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 6 \tab 2 \tab 5 \tab 1 \tab 2 \tab 9 \tab 4 \tab 1 \tab 1 \tab 18 \tab 3 \tab - \tab 1 \tab 13 \tab 8 \tab 2 \tab 3 \tab 1 \tab 5 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 5 \tab 5 \tab 13 \tab 7 \tab 1 \tab 19 \tab 7 \tab 1 \tab 27 \tab 2 \tab 1 \tab 5 \tab 6 \tab 2 \tab - \tab 2 \tab - \tab 5 \tab 2 \tab 2 \tab - \tab 20 \tab - \tab 2 \tab - \tab - \tab 11 \tab 3 \tab - \tab 5 \tab 7 \tab 2 \tab - \tab 14 \tab 9 \tab 5 \tab 36 \tab 2 \tab 1 \tab - \tab 11 \tab 2 \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 4 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 8 \tab 3 \tab 1 \tab 4 \tab 1 \tab 4 \tab 6 \tab 20 \tab 1 \tab 4 \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab 1 \tab 4 \tab 4 \tab 4 \tab 3 \tab 1 \tab 6 \tab 1 \tab - \tab 1 \tab 2 \tab 4 \tab - \tab 3 \tab - \tab 2 \tab 3 \tab 4 \tab 8 \tab 49 \tab 12 \tab 1 \tab 1 \tab 3 \tab 1 \tab 7 \tab 1 \tab 7 \tab 6 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 20 \tab 3 \tab 2 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab - \tab 3 \tab 12 \tab - \tab 6 \tab 18 \tab - \tab 13 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab 4 \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 793 \tab 12 \tab 2 \tab - \tab - \tab 2 \tab - \tab 5 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 5 \tab 2 \tab 12 \tab - \tab - \tab 4 \tab 3 \tab 1 \tab 1 \tab 14 \tab 9 \tab - \tab 1 \tab 6 \tab 12 \tab 7 \tab 1 \tab 2 \tab 4 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 6 \tab 7 \tab 14 \tab 5 \tab - \tab 5 \tab 7 \tab 2 \tab 22 \tab - \tab 4 \tab 4 \tab 10 \tab 5 \tab 2 \tab 4 \tab 2 \tab 4 \tab 4 \tab - \tab - \tab 20 \tab 1 \tab 3 \tab - \tab 1 \tab 14 \tab 2 \tab 1 \tab 3 \tab 8 \tab 1 \tab 3 \tab 18 \tab 13 \tab 7 \tab 38 \tab - \tab 1 \tab - \tab 5 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 4 \tab 10 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 12 \tab 1 \tab - \tab 11 \tab 2 \tab 2 \tab - \tab 27 \tab 6 \tab 7 \tab 2 \tab 2 \tab 1 \tab - \tab 5 \tab 1 \tab 3 \tab 2 \tab 5 \tab 1 \tab - \tab 3 \tab 4 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 4 \tab 16 \tab 63 \tab 15 \tab 2 \tab 1 \tab 7 \tab 4 \tab 6 \tab 1 \tab 3 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 17 \tab 3 \tab 5 \tab 5 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 5 \tab 7 \tab 1 \tab 3 \tab 16 \tab - \tab 4 \tab 4 \tab 2 \tab 5 \tab - \tab 1 \tab 4 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab 3 \tab - \tab 1 \tab 10 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 7 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1541 \tab 13 \tab 2 \tab 3 \tab 2 \tab 7 \tab 1 \tab 3 \tab 1 \tab 7 \tab 3 \tab 1 \tab - \tab 1 \tab 7 \tab 4 \tab 13 \tab 3 \tab 2 \tab 18 \tab 2 \tab 1 \tab 2 \tab 21 \tab 12 \tab - \tab 1 \tab 16 \tab 27 \tab 9 \tab 5 \tab 2 \tab 9 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 20 \tab 22 \tab 7 \tab 1 \tab 19 \tab 14 \tab 2 \tab 43 \tab 7 \tab 9 \tab 7 \tab 12 \tab 8 \tab - \tab 4 \tab 5 \tab 2 \tab 12 \tab 3 \tab - \tab 17 \tab 1 \tab 8 \tab - \tab - \tab 19 \tab 7 \tab - \tab 5 \tab 15 \tab 6 \tab 8 \tab 32 \tab 12 \tab 5 \tab 70 \tab - \tab 1 \tab 1 \tab 12 \tab 9 \tab 7 \tab 1 \tab 4 \tab 4 \tab 1 \tab 3 \tab 4 \tab 4 \tab 6 \tab 12 \tab 10 \tab 1 \tab 3 \tab 7 \tab 2 \tab 5 \tab 4 \tab 17 \tab 4 \tab 2 \tab 5 \tab 1 \tab 11 \tab 15 \tab 49 \tab 5 \tab 14 \tab 2 \tab 2 \tab 5 \tab 2 \tab 7 \tab 2 \tab 5 \tab 4 \tab 3 \tab 1 \tab - \tab 1 \tab 3 \tab 5 \tab 2 \tab 6 \tab 3 \tab 1 \tab 7 \tab 2 \tab 3 \tab 4 \tab 17 \tab 13 \tab 78 \tab 30 \tab 14 \tab 2 \tab 13 \tab 6 \tab 10 \tab 2 \tab 11 \tab 4 \tab - \tab 1 \tab 4 \tab - \tab 2 \tab 30 \tab 20 \tab 1 \tab 3 \tab - \tab 7 \tab 2 \tab - \tab - \tab 11 \tab 15 \tab 1 \tab 3 \tab 37 \tab - \tab 9 \tab 3 \tab 5 \tab 12 \tab - \tab 1 \tab 17 \tab 6 \tab 2 \tab 9 \tab 5 \tab - \tab 5 \tab 1 \tab - \tab 7 \tab - \tab - \tab 4 \tab - \tab 1 \tab 4 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab 2 \tab - \tab - \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 4 \tab 1 \tab - \tab - \tab - \tab 1 \tab 6 \tab 1 \tab 1 \tab - \tab 1 \tab 5 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1580 \tab 11 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 7 \tab 2 \tab 9 \tab - \tab 1 \tab 4 \tab 4 \tab 10 \tab 4 \tab 15 \tab 1 \tab 2 \tab 13 \tab - \tab - \tab - \tab 16 \tab 10 \tab 1 \tab 7 \tab 33 \tab 15 \tab 12 \tab 8 \tab 4 \tab 8 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab 5 \tab 7 \tab 25 \tab 11 \tab 1 \tab 20 \tab 11 \tab 3 \tab 19 \tab 6 \tab 11 \tab 5 \tab 5 \tab 4 \tab 1 \tab 4 \tab 1 \tab 13 \tab 3 \tab 1 \tab 2 \tab 19 \tab 2 \tab 2 \tab 2 \tab - \tab 21 \tab 7 \tab 1 \tab 4 \tab 9 \tab 5 \tab 4 \tab 27 \tab 9 \tab 9 \tab 52 \tab 4 \tab 3 \tab - \tab 28 \tab 9 \tab 7 \tab 2 \tab 3 \tab 4 \tab 2 \tab 1 \tab 4 \tab 1 \tab 5 \tab 11 \tab 10 \tab - \tab 1 \tab 5 \tab - \tab 2 \tab 5 \tab 16 \tab 3 \tab 2 \tab 3 \tab 2 \tab 6 \tab 7 \tab 41 \tab 3 \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab 5 \tab - \tab 8 \tab 2 \tab 2 \tab 3 \tab 1 \tab 7 \tab 5 \tab 3 \tab 1 \tab 8 \tab 3 \tab 2 \tab 5 \tab 1 \tab 1 \tab 3 \tab 9 \tab 18 \tab 115 \tab 5 \tab 8 \tab 9 \tab 10 \tab 8 \tab 7 \tab 3 \tab 6 \tab 12 \tab - \tab - \tab 3 \tab - \tab 1 \tab 25 \tab 5 \tab 6 \tab 7 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 7 \tab 17 \tab - \tab 3 \tab 32 \tab - \tab 14 \tab 3 \tab 5 \tab 5 \tab 2 \tab - \tab 10 \tab 4 \tab 6 \tab 14 \tab - \tab 2 \tab 5 \tab 1 \tab - \tab 12 \tab - \tab 1 \tab 9 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 2 \tab 3 \tab 3 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1556 \tab 4 \tab 1 \tab 2 \tab 1 \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 16 \tab 1 \tab 17 \tab - \tab 2 \tab 12 \tab 7 \tab 1 \tab - \tab 22 \tab 15 \tab - \tab 1 \tab 13 \tab 16 \tab 10 \tab 4 \tab 3 \tab 4 \tab 6 \tab - \tab 2 \tab 4 \tab 2 \tab 6 \tab 8 \tab 26 \tab 6 \tab 1 \tab 17 \tab 12 \tab 4 \tab 33 \tab 3 \tab 6 \tab 4 \tab 10 \tab 8 \tab 2 \tab 1 \tab 2 \tab 8 \tab 9 \tab - \tab 3 \tab 25 \tab 1 \tab 6 \tab 1 \tab 2 \tab 20 \tab - \tab 2 \tab 2 \tab 16 \tab 7 \tab 2 \tab 13 \tab 10 \tab 13 \tab 44 \tab 2 \tab 2 \tab - \tab 27 \tab 5 \tab 8 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 4 \tab - \tab - \tab 6 \tab 12 \tab 1 \tab 3 \tab 4 \tab 2 \tab 1 \tab 4 \tab 14 \tab 2 \tab 2 \tab 2 \tab - \tab 5 \tab 9 \tab 46 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab 3 \tab 5 \tab 1 \tab 8 \tab 5 \tab 3 \tab 2 \tab 3 \tab 6 \tab 1 \tab 1 \tab 4 \tab 9 \tab 3 \tab 1 \tab 3 \tab 1 \tab 2 \tab 2 \tab 9 \tab 18 \tab 120 \tab 12 \tab 13 \tab 2 \tab 5 \tab 5 \tab 9 \tab 5 \tab 9 \tab 9 \tab - \tab - \tab 4 \tab - \tab - \tab 19 \tab 10 \tab 6 \tab 5 \tab - \tab 10 \tab 1 \tab - \tab - \tab 10 \tab 12 \tab 1 \tab 1 \tab 31 \tab - \tab 15 \tab 4 \tab 5 \tab 10 \tab 2 \tab - \tab 10 \tab 6 \tab 5 \tab 8 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 8 \tab 2 \tab 1 \tab 6 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab 4 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 6 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1646 \tab 4 \tab - \tab 1 \tab - \tab 3 \tab - \tab 6 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 9 \tab 2 \tab 25 \tab 2 \tab - \tab 12 \tab 11 \tab 1 \tab 3 \tab 15 \tab 12 \tab 1 \tab 5 \tab 18 \tab 16 \tab 11 \tab 3 \tab - \tab 5 \tab 1 \tab 1 \tab 2 \tab 7 \tab 1 \tab 1 \tab 10 \tab 40 \tab 8 \tab 5 \tab 15 \tab 7 \tab 6 \tab 16 \tab 1 \tab 5 \tab 5 \tab 11 \tab 5 \tab 2 \tab 3 \tab 2 \tab 3 \tab 3 \tab 1 \tab 3 \tab 28 \tab 2 \tab 5 \tab - \tab - \tab 21 \tab 2 \tab 1 \tab 6 \tab 13 \tab 2 \tab 1 \tab 20 \tab 7 \tab 11 \tab 37 \tab 2 \tab - \tab 1 \tab 15 \tab 2 \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 5 \tab 2 \tab - \tab 5 \tab 9 \tab 3 \tab - \tab 5 \tab 3 \tab 5 \tab - \tab 20 \tab 1 \tab 1 \tab 1 \tab - \tab 7 \tab 9 \tab 42 \tab 4 \tab - \tab 3 \tab - \tab 1 \tab - \tab 16 \tab - \tab 3 \tab 1 \tab 3 \tab 1 \tab 3 \tab 4 \tab 5 \tab - \tab 4 \tab 6 \tab 4 \tab - \tab 6 \tab 4 \tab 1 \tab 4 \tab 9 \tab 15 \tab 100 \tab 4 \tab 6 \tab 1 \tab 7 \tab 3 \tab 4 \tab 2 \tab 11 \tab 4 \tab - \tab - \tab - \tab 2 \tab 1 \tab 18 \tab 5 \tab 4 \tab 2 \tab 1 \tab 5 \tab 3 \tab - \tab 1 \tab 3 \tab 19 \tab 1 \tab 2 \tab 37 \tab - \tab 13 \tab 4 \tab 2 \tab 14 \tab 1 \tab - \tab 12 \tab 1 \tab - \tab 9 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab 11 \tab 1 \tab - \tab 7 \tab - \tab 3 \tab 3 \tab - \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 7 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1573 \tab 4 \tab - \tab 1 \tab 1 \tab 9 \tab - \tab 4 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 11 \tab 3 \tab 15 \tab 1 \tab 2 \tab 4 \tab 12 \tab - \tab 2 \tab 13 \tab 11 \tab - \tab 5 \tab 15 \tab 17 \tab 7 \tab 3 \tab 1 \tab 5 \tab - \tab - \tab 2 \tab - \tab 1 \tab 4 \tab 14 \tab 18 \tab 4 \tab 1 \tab 14 \tab 11 \tab 2 \tab 33 \tab 6 \tab 4 \tab 3 \tab 8 \tab 3 \tab - \tab - \tab 4 \tab 8 \tab 5 \tab - \tab - \tab 19 \tab 3 \tab 7 \tab 5 \tab 2 \tab 13 \tab 2 \tab 2 \tab 3 \tab 20 \tab 4 \tab 2 \tab 20 \tab 4 \tab 8 \tab 39 \tab 3 \tab 4 \tab 2 \tab 10 \tab 4 \tab 10 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 2 \tab 4 \tab 6 \tab 10 \tab 1 \tab 7 \tab 4 \tab 4 \tab 1 \tab 5 \tab 15 \tab 3 \tab 1 \tab 7 \tab - \tab 6 \tab 5 \tab 31 \tab - \tab 5 \tab 2 \tab 5 \tab 1 \tab - \tab 6 \tab 1 \tab 9 \tab 5 \tab 2 \tab 6 \tab 3 \tab 6 \tab 5 \tab 3 \tab 1 \tab 8 \tab 5 \tab - \tab 11 \tab 4 \tab 3 \tab 4 \tab 9 \tab 19 \tab 101 \tab 10 \tab 6 \tab 5 \tab 2 \tab 3 \tab 4 \tab 7 \tab 9 \tab 7 \tab - \tab - \tab 3 \tab - \tab 1 \tab 24 \tab 9 \tab 6 \tab 3 \tab 1 \tab 13 \tab 4 \tab - \tab 1 \tab 3 \tab 13 \tab 2 \tab 2 \tab 32 \tab - \tab 12 \tab 6 \tab 5 \tab 16 \tab 1 \tab 1 \tab 15 \tab 7 \tab 6 \tab 3 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 10 \tab - \tab 1 \tab 9 \tab 2 \tab - \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab 7 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1459 \tab 4 \tab - \tab 1 \tab - \tab 10 \tab - \tab 6 \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab 15 \tab - \tab 12 \tab - \tab - \tab 7 \tab 4 \tab - \tab 1 \tab 22 \tab 14 \tab - \tab 3 \tab 19 \tab 19 \tab 10 \tab 3 \tab - \tab 5 \tab 2 \tab - \tab - \tab 4 \tab 2 \tab 1 \tab 13 \tab 34 \tab 5 \tab - \tab 15 \tab 4 \tab 9 \tab 26 \tab 6 \tab 6 \tab 7 \tab 14 \tab 4 \tab 2 \tab 3 \tab - \tab 13 \tab 11 \tab - \tab 1 \tab 25 \tab 2 \tab 4 \tab - \tab 1 \tab 15 \tab 4 \tab 1 \tab 3 \tab 10 \tab 5 \tab 8 \tab 12 \tab 6 \tab 5 \tab 30 \tab 4 \tab - \tab 1 \tab 23 \tab 7 \tab 9 \tab 1 \tab 3 \tab 2 \tab - \tab 4 \tab 1 \tab 2 \tab 4 \tab 2 \tab 13 \tab 1 \tab 4 \tab 6 \tab - \tab 1 \tab 2 \tab 13 \tab 3 \tab 2 \tab 5 \tab - \tab 8 \tab 8 \tab 50 \tab - \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab 4 \tab 1 \tab 4 \tab 5 \tab 2 \tab 1 \tab 1 \tab 6 \tab 4 \tab 3 \tab 2 \tab 10 \tab 7 \tab 1 \tab 5 \tab - \tab 2 \tab 3 \tab 16 \tab 20 \tab 122 \tab 20 \tab 12 \tab 2 \tab 9 \tab 5 \tab 14 \tab 11 \tab 13 \tab 19 \tab - \tab - \tab 2 \tab 1 \tab - \tab 28 \tab 13 \tab 13 \tab 4 \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 2 \tab 23 \tab 1 \tab 2 \tab 45 \tab - \tab 12 \tab 4 \tab 7 \tab 15 \tab 1 \tab - \tab 12 \tab 3 \tab 4 \tab 15 \tab 3 \tab 1 \tab 3 \tab 3 \tab 3 \tab 11 \tab 2 \tab 1 \tab 13 \tab - \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 4 \tab - \tab 2 \tab 3 \tab 7 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab 1 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2533 \tab 7 \tab - \tab 1 \tab 2 \tab 9 \tab - \tab 15 \tab - \tab 4 \tab 1 \tab 3 \tab 1 \tab - \tab 13 \tab 12 \tab 24 \tab 1 \tab 1 \tab 29 \tab 19 \tab - \tab 2 \tab 31 \tab 34 \tab - \tab 7 \tab 31 \tab 22 \tab 12 \tab 6 \tab 2 \tab 8 \tab 5 \tab 2 \tab 4 \tab 7 \tab - \tab 14 \tab 21 \tab 50 \tab 6 \tab 1 \tab 28 \tab 9 \tab 8 \tab 42 \tab 3 \tab 9 \tab 14 \tab 24 \tab 4 \tab 5 \tab 1 \tab 3 \tab 14 \tab 20 \tab - \tab - \tab 39 \tab - \tab 4 \tab - \tab 2 \tab 18 \tab 6 \tab 2 \tab 10 \tab 19 \tab 5 \tab 8 \tab 23 \tab 13 \tab 15 \tab 48 \tab 3 \tab 1 \tab 2 \tab 31 \tab 11 \tab 10 \tab 2 \tab 6 \tab 2 \tab - \tab 4 \tab 3 \tab - \tab 2 \tab 7 \tab 13 \tab - \tab 4 \tab 7 \tab 3 \tab 2 \tab 3 \tab 23 \tab 7 \tab 3 \tab 8 \tab 1 \tab 31 \tab 20 \tab 42 \tab 5 \tab 10 \tab 1 \tab 4 \tab 4 \tab 7 \tab 14 \tab 1 \tab 5 \tab 3 \tab 3 \tab 6 \tab 2 \tab 7 \tab 10 \tab 1 \tab 1 \tab 9 \tab 8 \tab 1 \tab 11 \tab 1 \tab 5 \tab - \tab 13 \tab 28 \tab 138 \tab 23 \tab 10 \tab 7 \tab 7 \tab 7 \tab 10 \tab 16 \tab 12 \tab 4 \tab - \tab - \tab 3 \tab - \tab 2 \tab 51 \tab 11 \tab 11 \tab 9 \tab 1 \tab 3 \tab 4 \tab 1 \tab 1 \tab 13 \tab 22 \tab 3 \tab 6 \tab 50 \tab 1 \tab 42 \tab 7 \tab 6 \tab 28 \tab 1 \tab - \tab 14 \tab 6 \tab 8 \tab 11 \tab 5 \tab 2 \tab 6 \tab 3 \tab 1 \tab 7 \tab 1 \tab - \tab 16 \tab - \tab 2 \tab 1 \tab 6 \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 4 \tab 4 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab 13 \tab - \tab 7 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 6 \tab - \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 3 \tab 6 \tab - \tab 5 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 1121 \tab 3 \tab - \tab - \tab 1 \tab 7 \tab - \tab 3 \tab - \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 4 \tab 5 \tab - \tab - \tab 8 \tab 11 \tab - \tab 1 \tab 13 \tab 11 \tab - \tab 1 \tab 12 \tab 16 \tab 4 \tab 2 \tab 1 \tab 4 \tab 2 \tab 2 \tab 2 \tab 5 \tab 3 \tab 2 \tab 11 \tab 20 \tab 4 \tab 1 \tab 13 \tab 6 \tab 1 \tab 21 \tab 1 \tab 3 \tab - \tab 9 \tab 3 \tab 4 \tab - \tab 1 \tab 6 \tab 7 \tab - \tab 1 \tab 16 \tab - \tab 4 \tab 1 \tab - \tab 13 \tab 4 \tab 1 \tab 1 \tab 11 \tab 2 \tab 5 \tab 17 \tab 7 \tab 3 \tab 32 \tab 6 \tab 2 \tab - \tab 7 \tab 1 \tab 2 \tab 1 \tab 5 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 6 \tab 1 \tab 3 \tab 8 \tab 1 \tab 1 \tab 3 \tab 8 \tab 6 \tab 3 \tab 1 \tab - \tab 9 \tab 8 \tab 20 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab 6 \tab 3 \tab 1 \tab 8 \tab 2 \tab 1 \tab 1 \tab 9 \tab 17 \tab 76 \tab 8 \tab 5 \tab 6 \tab 2 \tab 9 \tab 5 \tab 3 \tab 10 \tab 5 \tab - \tab - \tab - \tab 3 \tab 3 \tab 17 \tab 4 \tab 8 \tab 5 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 6 \tab 8 \tab 1 \tab 2 \tab 22 \tab - \tab 16 \tab 2 \tab 1 \tab 7 \tab - \tab - \tab 3 \tab 1 \tab 3 \tab 6 \tab - \tab - \tab 3 \tab 1 \tab - \tab 8 \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 5 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 7 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2023 \cr
   2010 \tab 1232 \tab 3 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab 4 \tab 3 \tab 7 \tab 1 \tab 2 \tab 8 \tab 7 \tab - \tab 1 \tab 16 \tab 9 \tab 1 \tab 6 \tab 14 \tab 13 \tab 5 \tab 2 \tab 3 \tab 2 \tab 2 \tab - \tab 1 \tab 4 \tab 4 \tab 3 \tab 10 \tab 17 \tab 4 \tab 1 \tab 13 \tab 5 \tab 3 \tab 17 \tab 5 \tab 4 \tab 3 \tab 5 \tab 3 \tab 2 \tab 1 \tab - \tab 9 \tab 6 \tab - \tab - \tab 17 \tab 1 \tab 2 \tab - \tab - \tab 6 \tab - \tab 1 \tab 1 \tab 11 \tab 2 \tab 2 \tab 8 \tab 6 \tab 6 \tab 20 \tab 2 \tab - \tab 1 \tab 16 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 6 \tab - \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 12 \tab 1 \tab 2 \tab 1 \tab 1 \tab 6 \tab 14 \tab 13 \tab 1 \tab 4 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab 9 \tab - \tab 2 \tab 2 \tab 9 \tab 12 \tab 71 \tab 8 \tab 11 \tab 5 \tab 3 \tab 9 \tab 6 \tab 1 \tab 6 \tab 1 \tab - \tab - \tab 3 \tab - \tab 3 \tab 20 \tab 6 \tab 7 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab 3 \tab 9 \tab - \tab 1 \tab 22 \tab - \tab 9 \tab 3 \tab 1 \tab 11 \tab - \tab - \tab 7 \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 6 \tab 1 \tab 2 \tab 6 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 57186 \tab 89 \tab 10 \tab 13 \tab 12 \tab 66 \tab 2 \tab 66 \tab 9 \tab 46 \tab 13 \tab 15 \tab 24 \tab 13 \tab 122 \tab 43 \tab 168 \tab 14 \tab 17 \tab 140 \tab 100 \tab 8 \tab 16 \tab 230 \tab 166 \tab 4 \tab 42 \tab 227 \tab 207 \tab 108 \tab 49 \tab 22 \tab 74 \tab 23 \tab 15 \tab 26 \tab 50 \tab 16 \tab 55 \tab 139 \tab 305 \tab 82 \tab 15 \tab 209 \tab 113 \tab 45 \tab 345 \tab 51 \tab 77 \tab 75 \tab 151 \tab 53 \tab 24 \tab 33 \tab 28 \tab 102 \tab 96 \tab 9 \tab 14 \tab 290 \tab 15 \tab 56 \tab 12 \tab 12 \tab 225 \tab 41 \tab 14 \tab 50 \tab 162 \tab 45 \tab 48 \tab 237 \tab 108 \tab 99 \tab 574 \tab 33 \tab 19 \tab 11 \tab 219 \tab 64 \tab 78 \tab 15 \tab 33 \tab 30 \tab 7 \tab 25 \tab 40 \tab 19 \tab 29 \tab 73 \tab 116 \tab 12 \tab 34 \tab 54 \tab 24 \tab 34 \tab 35 \tab 188 \tab 38 \tab 24 \tab 56 \tab 11 \tab 111 \tab 105 \tab 452 \tab 29 \tab 62 \tab 16 \tab 28 \tab 38 \tab 21 \tab 73 \tab 13 \tab 65 \tab 45 \tab 35 \tab 30 \tab 21 \tab 58 \tab 51 \tab 20 \tab 21 \tab 77 \tab 51 \tab 10 \tab 76 \tab 19 \tab 30 \tab 32 \tab 128 \tab 205 \tab 1167 \tab 164 \tab 106 \tab 53 \tab 86 \tab 68 \tab 88 \tab 55 \tab 107 \tab 88 \tab 1 \tab 4 \tab 33 \tab 8 \tab 19 \tab 297 \tab 102 \tab 83 \tab 56 \tab 8 \tab 60 \tab 33 \tab 4 \tab 10 \tab 74 \tab 173 \tab 12 \tab 40 \tab 383 \tab 4 \tab 177 \tab 49 \tab 43 \tab 138 \tab 10 \tab 5 \tab 115 \tab 47 \tab 46 \tab 94 \tab 26 \tab 9 \tab 38 \tab 20 \tab 7 \tab 115 \tab 10 \tab 8 \tab 91 \tab 13 \tab 11 \tab 19 \tab 24 \tab 8 \tab 5 \tab 21 \tab 4 \tab 20 \tab 4 \tab 6 \tab 4 \tab 7 \tab 25 \tab 6 \tab 8 \tab 14 \tab 4 \tab 2 \tab 7 \tab 11 \tab 7 \tab 29 \tab 14 \tab 4 \tab 1 \tab 11 \tab 12 \tab 6 \tab 2 \tab 11 \tab 9 \tab 6 \tab 28 \tab 11 \tab 21 \tab 22 \tab 30 \tab 18 \tab 11 \tab 3 \tab 24 \tab 3 \tab 10 \tab 55 \tab 3 \tab 31 \tab 6 \tab 3 \tab 20 \tab 6 \tab 8 \tab 2 \tab 3 \tab 4 \tab 16 \tab 2 \tab 6 \tab 17 \tab 5 \tab 17 \tab 8 \tab 23 \tab 15 \tab 21 \tab 5 \tab 7 \tab 3 \tab 10 \tab 23 \tab 7 \tab 23 \tab 6 \tab 7 \tab 21 \tab 3 \tab 10 \tab 3 \tab 1 \tab 3 \tab 3 \tab 5 \tab 18 \tab 5 \tab 9 \tab 8 \tab 10 \tab 9 \tab 3 \tab 10 \tab 3 \tab 3 \tab 3 \tab 5 \tab 3 \tab 5 \tab 5 \tab 1 \tab 5 \tab 14 \tab 24 \tab 2 \tab 2 \tab 4 \tab 4 \tab 10 \tab 8 \tab 10 \tab 8 \tab 2 \tab 11 \tab 21 \tab 1 \tab 15 \tab 1 \tab 2 \tab 11 \tab 1 \tab 4 \tab 2 \tab 5 \tab 3 \tab 2 \tab 4 \tab 4 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 4 \tab 4 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} i don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
