% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_598.R
\name{spsei10educ}
\alias{spsei10educ}
\title{spsei10educ}
\source{
General Social Survey https://gss.norc.org
}
\description{
Percentage of some college educ in spocc10 based on acs 2010
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 11.9 \tab 12.2 \tab 12.6 \tab 13.1 \tab 13.3 \tab 13.5 \tab 14.1 \tab 14.2 \tab 15.5 \tab 16 \tab 16.5 \tab 16.7 \tab 16.8 \tab 16.9 \tab 17.2 \tab 17.3 \tab 17.5 \tab 17.7 \tab 17.9 \tab 18 \tab 18.6 \tab 18.7 \tab 19.2 \tab 19.3 \tab 19.8 \tab 20.4 \tab 20.6 \tab 21 \tab 21.6 \tab 21.7 \tab 22 \tab 22.3 \tab 22.5 \tab 22.7 \tab 22.9 \tab 23.6 \tab 23.9 \tab 24.1 \tab 24.3 \tab 25.1 \tab 25.4 \tab 25.6 \tab 25.7 \tab 26.2 \tab 26.6 \tab 27.1 \tab 27.5 \tab 27.8 \tab 28 \tab 28.2 \tab 28.3 \tab 29.1 \tab 29.5 \tab 29.7 \tab 30.7 \tab 31 \tab 31.2 \tab 31.9 \tab 32.1 \tab 32.4 \tab 32.5 \tab 33.7 \tab 33.9 \tab 34.2 \tab 34.3 \tab 34.8 \tab 35 \tab 35.1 \tab 36 \tab 36.2 \tab 36.5 \tab 37 \tab 37.4 \tab 37.7 \tab 37.8 \tab 38.5 \tab 39.2 \tab 39.6 \tab 40.2 \tab 40.4 \tab 40.8 \tab 40.9 \tab 41.6 \tab 41.8 \tab 42 \tab 42.3 \tab 42.8 \tab 43.7 \tab 43.9 \tab 44 \tab 44.2 \tab 45.2 \tab 45.4 \tab 45.5 \tab 46.1 \tab 46.3 \tab 46.4 \tab 47.9 \tab 48.5 \tab 48.8 \tab 49 \tab 49.3 \tab 50 \tab 50.1 \tab 50.6 \tab 50.7 \tab 51.1 \tab 51.7 \tab 52.1 \tab 53 \tab 53.9 \tab 54.2 \tab 54.6 \tab 54.8 \tab 55.2 \tab 55.4 \tab 55.5 \tab 55.6 \tab 55.7 \tab 55.8 \tab 55.9 \tab 56.1 \tab 56.7 \tab 57.1 \tab 57.6 \tab 57.7 \tab 58.5 \tab 58.7 \tab 58.8 \tab 58.9 \tab 59.2 \tab 62.4 \tab 62.7 \tab 63.1 \tab 63.9 \tab 64.1 \tab 64.6 \tab 65.1 \tab 65.3 \tab 65.5 \tab 65.6 \tab 66.8 \tab 66.9 \tab 67.3 \tab 68.6 \tab 69.1 \tab 69.8 \tab 69.9 \tab 70.1 \tab 71.6 \tab 71.7 \tab 72.2 \tab 73 \tab 73.3 \tab 73.9 \tab 75.5 \tab 75.6 \tab 75.9 \tab 76.9 \tab 77.2 \tab 78 \tab 79.2 \tab 79.5 \tab 79.9 \tab 80.1 \tab 81.4 \tab 82.3 \tab 82.4 \tab 82.5 \tab 82.8 \tab 83 \tab 85.7 \tab 86.3 \tab 86.7 \tab 87 \tab 87.3 \tab 90.2 \tab 90.5 \tab 90.6 \tab 91.9 \tab 92 \tab 92.1 \tab 93.1 \tab 93.3 \tab 93.7 \tab 94 \tab 94.1 \tab 94.4 \tab 94.8 \tab 95.1 \tab 95.5 \tab 97.3 \tab 97.5 \tab 97.8 \tab 98.9 \tab 99 \tab 99.1 \tab 99.2 \tab 99.4 \tab 99.5 \tab 99.6 \tab 99.7 \tab iap \tab no answer \tab 12.3 \tab 17 \tab 19.7 \tab 20.5 \tab 20.8 \tab 21.1 \tab 21.5 \tab 21.9 \tab 23.2 \tab 24 \tab 26 \tab 26.7 \tab 29.4 \tab 30.4 \tab 30.6 \tab 32.8 \tab 33.8 \tab 35.5 \tab 35.9 \tab 38.3 \tab 38.4 \tab 40.3 \tab 42.6 \tab 44.4 \tab 46.6 \tab 47.3 \tab 48.9 \tab 54.7 \tab 61 \tab 63.3 \tab 66 \tab 68.4 \tab 69.7 \tab 70.4 \tab 74.2 \tab 75.1 \tab 76.5 \tab 78.2 \tab 78.3 \tab 78.4 \tab 81 \tab 81.1 \tab 83.3 \tab 83.7 \tab 85.9 \tab 86.8 \tab 88.3 \tab 89.4 \tab 89.8 \tab 91.4 \tab 91.7 \tab 95 \tab 95.3 \tab 95.6 \tab 96.8 \tab don't know \tab 12.4 \tab 14.3 \tab 19 \tab 19.6 \tab 24.4 \tab 28.7 \tab 35.8 \tab 46.9 \tab 47.4 \tab 48.2 \tab 61.1 \tab 67.4 \tab 72.5 \tab 74.4 \tab 77.5 \tab 79.6 \tab 81.9 \tab 83.1 \tab 83.5 \tab 90.9 \tab 91.2 \tab 91.3 \tab 92.4 \tab 95.7 \tab 98.3 \tab 28.9 \tab 31.1 \tab 38.6 \tab 47.5 \tab 48.1 \tab 51 \tab 58.6 \tab 59.3 \tab 64.7 \tab 67.6 \tab 70.3 \tab 79.4 \tab 81.3 \tab 85.6 \tab 88.2 \tab 17.8 \tab 27 \tab 34.1 \tab 35.3 \tab 42.1 \tab 45.1 \tab 45.7 \tab 53.3 \tab 54.3 \tab 60.9 \tab 62.2 \tab 66.3 \tab 85.4 \tab 96.4 \tab 98.5 \tab 22.4 \tab 39.9 \tab 52.2 \tab 61.3 \tab 68.8 \tab 78.1 \tab 78.8 \tab 92.9 \tab 99.3 \tab 19.1 \tab 24.7 \tab 33.5 \tab 35.2 \tab 69.6 \tab 73.5 \tab 85.3 \tab 90.8 \tab 96 \tab 25.8 \tab 70.9 \tab 84.5 \tab 91.1 \tab 92.6 \tab 96.5 \tab 98.7 \tab 13.4 \tab 23 \tab 47.8 \tab 48.3 \tab 60.7 \tab 62.8 \tab 68.2 \tab 79.8 \tab 96.3 \tab 19.9 \tab 49.7 \tab 56.6 \tab 80.7 \tab 84.4 \tab 74.6 \tab 84 \tab 84.6 \tab 97 \tab 52.7 \tab 66.7 \tab 63.7 \tab 80 \tab 62.3 \tab 63.5 \tab 55.1 \tab 84.8 \tab 34.9 \tab 34.7 \tab 9.8 \tab 58.2 \tab 86.2 \tab 10.9 \tab 87.5 \tab 62.9 \tab 84.7 \tab not available in this release \tab Total \cr
   1972 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 3 \tab 11 \tab 1 \tab 1 \tab 6 \tab 19 \tab 2 \tab 1 \tab 3 \tab 3 \tab 25 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 7 \tab 2 \tab 1 \tab 7 \tab 1 \tab 4 \tab 2 \tab 7 \tab 3 \tab 33 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 14 \tab 10 \tab 6 \tab 25 \tab 2 \tab 12 \tab 1 \tab 2 \tab 3 \tab 2 \tab 5 \tab 2 \tab 1 \tab 2 \tab 2 \tab 7 \tab 1 \tab 4 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 7 \tab 1 \tab 1 \tab 1 \tab 7 \tab 2 \tab 3 \tab 1 \tab 3 \tab 1 \tab 2 \tab 17 \tab 2 \tab 4 \tab 1 \tab 4 \tab 1 \tab 1 \tab 9 \tab 2 \tab 1 \tab 1 \tab 14 \tab 12 \tab 5 \tab 1 \tab 5 \tab 16 \tab 1 \tab 32 \tab 8 \tab 2 \tab 2 \tab 2 \tab 3 \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 8 \tab 5 \tab 3 \tab 7 \tab 5 \tab 1 \tab 10 \tab 1 \tab 1 \tab 1 \tab 12 \tab 1 \tab 1 \tab 3 \tab 3 \tab 19 \tab 40 \tab 8 \tab 3 \tab 70 \tab 2 \tab 1 \tab 1 \tab 4 \tab 2 \tab 3 \tab 4 \tab 1 \tab 1 \tab 4 \tab 3 \tab 4 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab 7 \tab 1 \tab 3 \tab 7 \tab 1 \tab 2 \tab 2 \tab 6 \tab 5 \tab 1 \tab 1 \tab 1 \tab 6 \tab 8 \tab 6 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 4 \tab 3 \tab 4 \tab 5 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 9 \tab 1 \tab 1 \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 11 \tab 28 \tab 9 \tab 8 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 695 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 7 \tab - \tab 1 \tab 4 \tab 29 \tab 1 \tab 1 \tab 4 \tab 1 \tab 13 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 6 \tab 2 \tab - \tab 4 \tab - \tab 1 \tab 2 \tab 13 \tab 3 \tab 21 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab 10 \tab 11 \tab 9 \tab 25 \tab 2 \tab 12 \tab - \tab - \tab 2 \tab - \tab 5 \tab 4 \tab 1 \tab - \tab - \tab 9 \tab 2 \tab 2 \tab 4 \tab 3 \tab 4 \tab - \tab - \tab - \tab 3 \tab 7 \tab - \tab 1 \tab 2 \tab 12 \tab 10 \tab 2 \tab 3 \tab 2 \tab 8 \tab 5 \tab 7 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 12 \tab - \tab 4 \tab 2 \tab 12 \tab 12 \tab 1 \tab - \tab 5 \tab 13 \tab - \tab 13 \tab 1 \tab - \tab 5 \tab 1 \tab 3 \tab - \tab 3 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 2 \tab 2 \tab 2 \tab 6 \tab 2 \tab 9 \tab 1 \tab - \tab - \tab 15 \tab 2 \tab - \tab 3 \tab 1 \tab 23 \tab 23 \tab 6 \tab 1 \tab 45 \tab 2 \tab 4 \tab - \tab 5 \tab - \tab 4 \tab 3 \tab 1 \tab 1 \tab 4 \tab - \tab 3 \tab 3 \tab 10 \tab - \tab 5 \tab 1 \tab 5 \tab 1 \tab 2 \tab 6 \tab - \tab 4 \tab 4 \tab - \tab 2 \tab 2 \tab 9 \tab - \tab 1 \tab 2 \tab - \tab 9 \tab 9 \tab 3 \tab 2 \tab 4 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab 3 \tab 1 \tab 4 \tab 9 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 5 \tab 1 \tab 2 \tab - \tab 6 \tab - \tab - \tab 4 \tab - \tab 4 \tab - \tab 2 \tab 2 \tab 13 \tab 25 \tab 14 \tab 8 \tab - \tab 7 \tab - \tab - \tab 7 \tab 619 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab 2 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 5 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab 6 \tab - \tab - \tab 2 \tab 26 \tab - \tab - \tab 3 \tab 4 \tab 12 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 10 \tab 6 \tab - \tab 8 \tab 1 \tab 3 \tab 2 \tab 14 \tab 6 \tab 19 \tab - \tab 4 \tab 1 \tab - \tab - \tab 10 \tab 7 \tab 6 \tab 17 \tab 1 \tab 16 \tab - \tab 1 \tab - \tab 1 \tab 6 \tab 2 \tab 4 \tab - \tab - \tab 6 \tab 2 \tab 2 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 10 \tab 12 \tab - \tab - \tab - \tab 14 \tab 10 \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab 5 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 7 \tab - \tab - \tab - \tab 21 \tab 15 \tab - \tab 1 \tab 7 \tab 14 \tab - \tab 12 \tab 6 \tab 1 \tab 4 \tab - \tab 5 \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab 5 \tab - \tab 3 \tab 7 \tab - \tab 4 \tab 3 \tab 1 \tab 5 \tab - \tab - \tab 1 \tab 17 \tab - \tab - \tab 6 \tab 4 \tab 29 \tab 30 \tab 6 \tab 1 \tab 51 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 4 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 4 \tab 10 \tab - \tab 6 \tab 1 \tab - \tab - \tab 3 \tab 8 \tab - \tab 4 \tab 4 \tab - \tab - \tab 2 \tab 13 \tab 2 \tab 1 \tab 2 \tab 1 \tab 4 \tab 11 \tab 9 \tab 3 \tab 7 \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 3 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 4 \tab 3 \tab 1 \tab 5 \tab - \tab 6 \tab - \tab - \tab 4 \tab 1 \tab 3 \tab - \tab - \tab 3 \tab 19 \tab 25 \tab 16 \tab 3 \tab 1 \tab 4 \tab - \tab - \tab 5 \tab 607 \tab - \tab - \tab 6 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 9 \tab 1 \tab 1 \tab 4 \tab 15 \tab - \tab 3 \tab 8 \tab 1 \tab 11 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab 9 \tab 2 \tab - \tab 8 \tab - \tab 2 \tab 4 \tab 9 \tab - \tab 19 \tab - \tab 3 \tab 8 \tab - \tab - \tab 10 \tab 7 \tab 8 \tab 23 \tab 1 \tab 23 \tab 4 \tab 3 \tab 3 \tab 1 \tab 6 \tab 1 \tab 4 \tab 1 \tab 1 \tab 9 \tab 1 \tab 3 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 8 \tab 6 \tab - \tab - \tab 1 \tab 10 \tab 3 \tab - \tab - \tab 7 \tab - \tab 4 \tab 5 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 15 \tab - \tab 4 \tab 3 \tab 11 \tab 7 \tab 3 \tab 1 \tab 5 \tab 10 \tab - \tab 20 \tab 4 \tab 1 \tab 3 \tab 1 \tab 4 \tab - \tab 3 \tab 4 \tab - \tab - \tab 2 \tab 2 \tab - \tab 7 \tab 5 \tab 2 \tab 2 \tab 1 \tab - \tab 6 \tab 2 \tab 1 \tab - \tab 18 \tab - \tab - \tab 4 \tab 5 \tab 15 \tab 33 \tab 6 \tab 3 \tab 41 \tab 4 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 3 \tab 1 \tab 2 \tab 6 \tab 2 \tab - \tab - \tab 7 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 5 \tab - \tab 3 \tab 7 \tab 1 \tab - \tab - \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab 9 \tab 5 \tab 3 \tab 7 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 5 \tab 3 \tab 4 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 8 \tab 3 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab 12 \tab 24 \tab 15 \tab 6 \tab - \tab 2 \tab 2 \tab - \tab 8 \tab 653 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 4 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1 \tab 4 \tab - \tab 1 \tab 3 \tab 5 \tab 10 \tab - \tab - \tab 2 \tab 19 \tab - \tab 1 \tab 6 \tab 3 \tab 9 \tab 2 \tab - \tab 3 \tab 2 \tab - \tab 3 \tab 5 \tab 1 \tab - \tab 3 \tab 6 \tab 5 \tab 3 \tab 8 \tab 3 \tab 31 \tab - \tab 2 \tab 4 \tab - \tab - \tab 4 \tab 10 \tab 10 \tab 20 \tab - \tab 18 \tab 1 \tab 1 \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 9 \tab 3 \tab 3 \tab 2 \tab - \tab 3 \tab - \tab 2 \tab - \tab 10 \tab 12 \tab - \tab - \tab - \tab 9 \tab 4 \tab 3 \tab - \tab 2 \tab 3 \tab 3 \tab 6 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 11 \tab - \tab 1 \tab - \tab 18 \tab 18 \tab - \tab 1 \tab 6 \tab 12 \tab 1 \tab 27 \tab 6 \tab 2 \tab 4 \tab 2 \tab 4 \tab - \tab 7 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 3 \tab 4 \tab - \tab 1 \tab 3 \tab 4 \tab 6 \tab - \tab 1 \tab - \tab 24 \tab 2 \tab - \tab 2 \tab 1 \tab 18 \tab 30 \tab 6 \tab 2 \tab 45 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 8 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 9 \tab 1 \tab 1 \tab 4 \tab 3 \tab - \tab - \tab 2 \tab - \tab 5 \tab 12 \tab 1 \tab 6 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 2 \tab 3 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 7 \tab 2 \tab 4 \tab - \tab 10 \tab 1 \tab 1 \tab 5 \tab 2 \tab 2 \tab - \tab 1 \tab 4 \tab 10 \tab 17 \tab 13 \tab 8 \tab - \tab 5 \tab - \tab - \tab 2 \tab 646 \tab - \tab - \tab 5 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1 \tab 4 \tab 3 \tab 1 \tab - \tab 5 \tab 5 \tab - \tab 2 \tab 2 \tab 18 \tab - \tab - \tab 2 \tab 1 \tab 17 \tab 1 \tab - \tab - \tab - \tab 1 \tab 3 \tab 12 \tab 3 \tab - \tab 5 \tab 1 \tab 3 \tab 1 \tab 14 \tab 3 \tab 19 \tab - \tab 1 \tab 3 \tab - \tab - \tab 7 \tab 7 \tab 4 \tab 37 \tab - \tab 11 \tab 1 \tab - \tab 5 \tab 2 \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 12 \tab 1 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab 5 \tab 4 \tab - \tab - \tab 3 \tab 6 \tab 7 \tab 3 \tab - \tab 3 \tab 4 \tab 4 \tab 3 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 10 \tab - \tab 2 \tab - \tab 13 \tab 16 \tab 3 \tab - \tab 5 \tab 14 \tab 1 \tab 18 \tab 6 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 6 \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 3 \tab 5 \tab - \tab 4 \tab 4 \tab 1 \tab 4 \tab - \tab - \tab - \tab 11 \tab 5 \tab - \tab 2 \tab - \tab 13 \tab 36 \tab 6 \tab 3 \tab 46 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab 8 \tab 1 \tab 7 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 4 \tab 7 \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab 5 \tab 9 \tab 3 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 5 \tab 1 \tab 5 \tab 5 \tab - \tab 2 \tab - \tab - \tab 2 \tab 4 \tab - \tab 4 \tab 4 \tab 1 \tab 1 \tab 8 \tab 1 \tab - \tab 6 \tab 1 \tab 5 \tab - \tab 2 \tab - \tab 12 \tab 17 \tab 14 \tab 5 \tab - \tab 2 \tab - \tab - \tab 3 \tab 733 \tab 1 \tab - \tab 3 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab 5 \tab - \tab 1 \tab 3 \tab 20 \tab - \tab - \tab 3 \tab - \tab 11 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 9 \tab - \tab - \tab 5 \tab - \tab 3 \tab 5 \tab 13 \tab 8 \tab 25 \tab - \tab 4 \tab 5 \tab - \tab - \tab 7 \tab 6 \tab 11 \tab 18 \tab - \tab 15 \tab 2 \tab 1 \tab 1 \tab - \tab 8 \tab 2 \tab 4 \tab 3 \tab - \tab 11 \tab 6 \tab 4 \tab 1 \tab 1 \tab 2 \tab 4 \tab - \tab 1 \tab 4 \tab 10 \tab 2 \tab - \tab 1 \tab 7 \tab 12 \tab 5 \tab 1 \tab 3 \tab 1 \tab 4 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 11 \tab - \tab 1 \tab 1 \tab 16 \tab 10 \tab - \tab 1 \tab 5 \tab 10 \tab - \tab 16 \tab 8 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 6 \tab 1 \tab - \tab 1 \tab 26 \tab 1 \tab - \tab 3 \tab 3 \tab 25 \tab 20 \tab 3 \tab - \tab 45 \tab - \tab 4 \tab - \tab 6 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 7 \tab 1 \tab 8 \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab - \tab 3 \tab 6 \tab 1 \tab 3 \tab - \tab 10 \tab - \tab 1 \tab 1 \tab - \tab 6 \tab 22 \tab 2 \tab 4 \tab - \tab 3 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 1 \tab 4 \tab 4 \tab 4 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 5 \tab 3 \tab 2 \tab - \tab 6 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 1 \tab 14 \tab 19 \tab 5 \tab 9 \tab - \tab 5 \tab 1 \tab 1 \tab 2 \tab 660 \tab - \tab - \tab 7 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 5 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 9 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 3 \tab 2 \tab - \tab 3 \tab 1 \tab 8 \tab - \tab 2 \tab - \tab 12 \tab - \tab - \tab 2 \tab - \tab 10 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 7 \tab 5 \tab 1 \tab 3 \tab 2 \tab 2 \tab 3 \tab 7 \tab 1 \tab 23 \tab - \tab 5 \tab 4 \tab - \tab 2 \tab 8 \tab 16 \tab 6 \tab 22 \tab - \tab 6 \tab - \tab - \tab 1 \tab - \tab 4 \tab 4 \tab 1 \tab - \tab - \tab 11 \tab 2 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 5 \tab 10 \tab - \tab 1 \tab 2 \tab 1 \tab 9 \tab 3 \tab 1 \tab - \tab 3 \tab - \tab 6 \tab - \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 14 \tab - \tab - \tab 1 \tab 19 \tab 13 \tab 1 \tab - \tab 3 \tab 5 \tab 1 \tab 14 \tab 6 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab 6 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 8 \tab 2 \tab - \tab 4 \tab 3 \tab 13 \tab 2 \tab - \tab - \tab 21 \tab 1 \tab - \tab 5 \tab 4 \tab 17 \tab 22 \tab 3 \tab 1 \tab 38 \tab - \tab 4 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 2 \tab 12 \tab - \tab 6 \tab - \tab 3 \tab 1 \tab 3 \tab 5 \tab 1 \tab 4 \tab 6 \tab 1 \tab 4 \tab - \tab 7 \tab 1 \tab - \tab 2 \tab 1 \tab 5 \tab 16 \tab 7 \tab 3 \tab 1 \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 6 \tab 3 \tab 1 \tab 2 \tab 3 \tab - \tab 2 \tab 2 \tab - \tab 6 \tab 1 \tab 2 \tab - \tab 4 \tab - \tab 1 \tab 3 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 8 \tab 19 \tab 12 \tab 6 \tab - \tab 5 \tab - \tab - \tab 1 \tab 686 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 5 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 7 \tab - \tab 1 \tab 4 \tab 17 \tab - \tab 3 \tab 4 \tab 3 \tab 12 \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 16 \tab 3 \tab - \tab 7 \tab - \tab 3 \tab 2 \tab 11 \tab 4 \tab 23 \tab - \tab 3 \tab 2 \tab - \tab 1 \tab 7 \tab 15 \tab 6 \tab 22 \tab 1 \tab 10 \tab 4 \tab 1 \tab 1 \tab 2 \tab 9 \tab 1 \tab 3 \tab 3 \tab - \tab 9 \tab 7 \tab 1 \tab 5 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 12 \tab 8 \tab - \tab - \tab 1 \tab 13 \tab 4 \tab 7 \tab - \tab 3 \tab 2 \tab 7 \tab 10 \tab 2 \tab 3 \tab - \tab 3 \tab - \tab - \tab 16 \tab - \tab 1 \tab 3 \tab 8 \tab 11 \tab 2 \tab - \tab 8 \tab 9 \tab 1 \tab 14 \tab 8 \tab 2 \tab 2 \tab - \tab 5 \tab - \tab 6 \tab 3 \tab 3 \tab - \tab 3 \tab 7 \tab 4 \tab 2 \tab 4 \tab - \tab 3 \tab 5 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab 22 \tab 2 \tab - \tab 4 \tab 5 \tab 18 \tab 20 \tab 3 \tab 2 \tab 37 \tab 3 \tab 7 \tab - \tab 5 \tab 1 \tab 2 \tab 3 \tab 1 \tab 4 \tab 3 \tab - \tab 2 \tab - \tab 14 \tab - \tab 4 \tab 1 \tab 2 \tab - \tab 5 \tab 2 \tab 1 \tab 2 \tab 4 \tab 1 \tab - \tab - \tab 9 \tab 1 \tab 2 \tab 1 \tab 1 \tab 8 \tab 12 \tab 4 \tab 4 \tab 6 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 6 \tab 2 \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab 2 \tab 2 \tab 9 \tab 4 \tab 1 \tab 1 \tab 6 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 10 \tab 25 \tab 9 \tab 8 \tab - \tab 4 \tab - \tab - \tab 8 \tab 962 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 8 \tab - \tab 1 \tab 1 \tab 17 \tab 1 \tab 1 \tab 4 \tab - \tab 8 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 4 \tab 3 \tab - \tab 7 \tab 1 \tab 2 \tab 1 \tab 6 \tab 2 \tab 18 \tab - \tab - \tab 2 \tab - \tab 1 \tab 9 \tab 5 \tab 9 \tab 25 \tab 1 \tab 11 \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab 8 \tab 6 \tab 2 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab - \tab 4 \tab 4 \tab - \tab - \tab 1 \tab 10 \tab 7 \tab 1 \tab - \tab 4 \tab 1 \tab 4 \tab 7 \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 16 \tab 1 \tab 1 \tab 4 \tab 9 \tab 16 \tab - \tab - \tab 5 \tab 8 \tab - \tab 18 \tab 6 \tab - \tab 1 \tab 2 \tab 4 \tab - \tab 4 \tab 1 \tab - \tab 2 \tab 2 \tab 4 \tab - \tab 6 \tab 6 \tab - \tab 2 \tab 5 \tab 3 \tab 2 \tab - \tab 2 \tab 1 \tab 22 \tab 2 \tab - \tab 4 \tab 3 \tab 18 \tab 29 \tab 5 \tab 3 \tab 51 \tab 1 \tab 4 \tab - \tab 3 \tab - \tab 5 \tab 6 \tab 2 \tab 2 \tab 5 \tab - \tab - \tab - \tab 8 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 7 \tab - \tab 2 \tab 2 \tab - \tab 5 \tab 1 \tab 14 \tab 1 \tab 1 \tab 1 \tab - \tab 8 \tab 15 \tab 7 \tab 9 \tab 4 \tab 5 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 5 \tab 7 \tab 2 \tab 6 \tab 7 \tab 1 \tab 2 \tab 2 \tab - \tab 3 \tab 3 \tab 2 \tab 5 \tab 3 \tab 1 \tab 1 \tab 10 \tab - \tab 1 \tab 4 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 17 \tab 9 \tab 12 \tab 9 \tab - \tab 6 \tab 1 \tab - \tab 7 \tab 722 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 3 \tab 3 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab 5 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 8 \tab - \tab - \tab 2 \tab 21 \tab - \tab 1 \tab 1 \tab 1 \tab 7 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab 6 \tab 1 \tab - \tab 3 \tab 2 \tab - \tab 4 \tab 8 \tab 3 \tab 20 \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab 9 \tab 8 \tab 8 \tab 16 \tab 1 \tab 7 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 7 \tab 3 \tab 2 \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 6 \tab 8 \tab - \tab - \tab 1 \tab 11 \tab 9 \tab 3 \tab - \tab 4 \tab 2 \tab 3 \tab 5 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 14 \tab - \tab - \tab 1 \tab 12 \tab 11 \tab - \tab - \tab 2 \tab 12 \tab - \tab 14 \tab 3 \tab 3 \tab 2 \tab 2 \tab 2 \tab - \tab 6 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 4 \tab 1 \tab 2 \tab 5 \tab - \tab 3 \tab 1 \tab - \tab - \tab 28 \tab 3 \tab - \tab 6 \tab 1 \tab 18 \tab 18 \tab 5 \tab 1 \tab 32 \tab - \tab 6 \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 5 \tab 1 \tab 1 \tab 1 \tab 13 \tab - \tab 5 \tab 2 \tab 3 \tab 1 \tab 4 \tab 5 \tab - \tab 5 \tab 11 \tab - \tab 2 \tab 3 \tab 7 \tab 1 \tab - \tab 6 \tab - \tab 1 \tab 14 \tab 3 \tab 4 \tab 3 \tab 4 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 7 \tab 5 \tab 5 \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 2 \tab 2 \tab 1 \tab 5 \tab - \tab 1 \tab 8 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 11 \tab 9 \tab 10 \tab 4 \tab - \tab 6 \tab - \tab - \tab - \tab 695 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 4 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 7 \tab 1 \tab - \tab - \tab 18 \tab - \tab 1 \tab 2 \tab 2 \tab 8 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 3 \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 5 \tab 7 \tab 3 \tab 16 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 9 \tab 7 \tab 6 \tab 20 \tab - \tab 9 \tab 3 \tab - \tab 2 \tab 1 \tab 4 \tab 1 \tab 4 \tab - \tab - \tab 5 \tab 2 \tab 2 \tab 3 \tab 2 \tab 4 \tab 4 \tab - \tab 1 \tab 9 \tab 5 \tab 1 \tab 1 \tab 2 \tab 7 \tab 9 \tab 1 \tab - \tab 3 \tab 1 \tab 3 \tab 7 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 12 \tab - \tab 1 \tab - \tab 15 \tab 12 \tab - \tab - \tab 6 \tab 13 \tab - \tab 14 \tab 10 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab 3 \tab 9 \tab - \tab 5 \tab 1 \tab 2 \tab 1 \tab 21 \tab 3 \tab - \tab 4 \tab - \tab 32 \tab 34 \tab 4 \tab 2 \tab 34 \tab - \tab 4 \tab - \tab 4 \tab - \tab 2 \tab 5 \tab 2 \tab 2 \tab 6 \tab - \tab 1 \tab 1 \tab 13 \tab - \tab 8 \tab - \tab - \tab - \tab 2 \tab 4 \tab - \tab 5 \tab 1 \tab - \tab - \tab 2 \tab 9 \tab 3 \tab - \tab - \tab 1 \tab 5 \tab 10 \tab 2 \tab 1 \tab 1 \tab 5 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab 9 \tab 3 \tab 6 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 3 \tab 2 \tab 4 \tab 9 \tab 1 \tab 2 \tab 1 \tab 11 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab 11 \tab 15 \tab 9 \tab 5 \tab 1 \tab 2 \tab - \tab - \tab 4 \tab 739 \tab - \tab 1 \tab 5 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 4 \tab 1 \tab 2 \tab 2 \tab - \tab 5 \tab - \tab 1 \tab 2 \tab 11 \tab - \tab 1 \tab 4 \tab 2 \tab 9 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 10 \tab 4 \tab - \tab 4 \tab 1 \tab 3 \tab 1 \tab 8 \tab 4 \tab 12 \tab - \tab 5 \tab 1 \tab - \tab 1 \tab 11 \tab 6 \tab 5 \tab 24 \tab - \tab 14 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab - \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 6 \tab 3 \tab 2 \tab - \tab 3 \tab 8 \tab 6 \tab 5 \tab - \tab 2 \tab - \tab 5 \tab 5 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 9 \tab 1 \tab 4 \tab - \tab 13 \tab 10 \tab 1 \tab 1 \tab 2 \tab 13 \tab - \tab 13 \tab 6 \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab 2 \tab 5 \tab 2 \tab 1 \tab 4 \tab 2 \tab 4 \tab - \tab 1 \tab - \tab 17 \tab 2 \tab - \tab 3 \tab 1 \tab 20 \tab 17 \tab 2 \tab 2 \tab 45 \tab 2 \tab 2 \tab - \tab 4 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 10 \tab - \tab 10 \tab - \tab 1 \tab 3 \tab 3 \tab 3 \tab - \tab 4 \tab 6 \tab - \tab 1 \tab 4 \tab 7 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 8 \tab 3 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 4 \tab 2 \tab 3 \tab 6 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 9 \tab 5 \tab 1 \tab - \tab 11 \tab - \tab - \tab 5 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 18 \tab 22 \tab 9 \tab 3 \tab 1 \tab 7 \tab - \tab - \tab 5 \tab 712 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab 1 \tab - \tab 5 \tab - \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 8 \tab - \tab 2 \tab 2 \tab 13 \tab - \tab 1 \tab 3 \tab 1 \tab 11 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab 11 \tab - \tab - \tab 7 \tab 1 \tab 3 \tab 3 \tab 6 \tab 2 \tab 20 \tab - \tab 5 \tab 3 \tab - \tab 1 \tab 10 \tab 14 \tab 4 \tab 21 \tab 1 \tab 11 \tab - \tab - \tab - \tab 2 \tab 1 \tab 4 \tab 3 \tab 1 \tab - \tab 9 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 2 \tab 5 \tab 10 \tab - \tab - \tab 3 \tab 12 \tab 6 \tab 2 \tab - \tab 4 \tab 1 \tab 4 \tab 6 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 13 \tab - \tab 1 \tab - \tab 11 \tab 9 \tab - \tab - \tab 2 \tab 6 \tab - \tab 20 \tab 10 \tab 4 \tab 2 \tab - \tab 2 \tab 2 \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab 4 \tab - \tab 1 \tab 6 \tab - \tab 2 \tab 3 \tab - \tab 6 \tab - \tab 1 \tab 2 \tab 31 \tab 2 \tab - \tab 5 \tab 3 \tab 26 \tab 15 \tab 4 \tab 2 \tab 25 \tab 1 \tab 6 \tab - \tab 4 \tab 4 \tab 3 \tab - \tab - \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 6 \tab - \tab 10 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab 3 \tab 7 \tab - \tab 1 \tab 1 \tab 6 \tab 1 \tab - \tab 1 \tab 1 \tab 5 \tab 15 \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab 2 \tab 7 \tab 1 \tab 11 \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 11 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab 4 \tab 9 \tab 20 \tab 13 \tab 7 \tab 1 \tab 3 \tab - \tab - \tab 6 \tab 1005 \tab - \tab 2 \tab 6 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 5 \tab 1 \tab - \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab 8 \tab - \tab 1 \tab - \tab - \tab 9 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 8 \tab 5 \tab - \tab 2 \tab 1 \tab 5 \tab 2 \tab 5 \tab 2 \tab 13 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab 8 \tab 11 \tab 8 \tab 19 \tab 2 \tab 12 \tab 1 \tab 2 \tab 2 \tab 1 \tab 6 \tab - \tab 1 \tab - \tab - \tab 4 \tab 5 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 6 \tab 4 \tab - \tab - \tab 1 \tab 7 \tab 9 \tab 4 \tab - \tab 6 \tab 1 \tab 4 \tab 9 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 9 \tab - \tab 1 \tab 3 \tab 6 \tab 10 \tab 1 \tab - \tab 7 \tab 10 \tab - \tab 8 \tab 8 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 4 \tab 1 \tab 3 \tab 5 \tab 2 \tab - \tab 3 \tab 1 \tab 5 \tab - \tab 1 \tab - \tab 19 \tab 2 \tab - \tab 1 \tab 5 \tab 20 \tab 17 \tab 3 \tab 1 \tab 35 \tab - \tab 4 \tab - \tab 3 \tab 5 \tab 4 \tab 3 \tab - \tab 1 \tab 2 \tab 2 \tab 2 \tab 2 \tab 9 \tab - \tab 10 \tab - \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 7 \tab 2 \tab - \tab 1 \tab 1 \tab 5 \tab 18 \tab 5 \tab 9 \tab 3 \tab 2 \tab 2 \tab - \tab 2 \tab 4 \tab 3 \tab 5 \tab 3 \tab 1 \tab 5 \tab 5 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 3 \tab 6 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 2 \tab 9 \tab 12 \tab 15 \tab 4 \tab - \tab 8 \tab - \tab - \tab 6 \tab 751 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab 4 \tab 2 \tab - \tab 1 \tab - \tab 3 \tab - \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 5 \tab 1 \tab - \tab 4 \tab 16 \tab - \tab 2 \tab 5 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 9 \tab 3 \tab 4 \tab 4 \tab 8 \tab 5 \tab 22 \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 7 \tab 15 \tab 8 \tab 18 \tab 1 \tab 11 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 6 \tab - \tab 4 \tab - \tab 2 \tab 7 \tab 1 \tab - \tab - \tab 3 \tab 6 \tab 1 \tab - \tab 2 \tab 11 \tab 5 \tab 5 \tab 1 \tab 3 \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab 11 \tab - \tab - \tab - \tab 14 \tab 8 \tab 1 \tab - \tab 1 \tab 10 \tab - \tab 13 \tab 7 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab 2 \tab 2 \tab - \tab 2 \tab 5 \tab - \tab 1 \tab 6 \tab - \tab - \tab 3 \tab 2 \tab 5 \tab 3 \tab - \tab - \tab 20 \tab 2 \tab - \tab 4 \tab 6 \tab 15 \tab 17 \tab 2 \tab 3 \tab 43 \tab 1 \tab 5 \tab - \tab 3 \tab 4 \tab 3 \tab 3 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab - \tab 13 \tab - \tab 9 \tab - \tab 1 \tab 1 \tab 6 \tab 4 \tab 1 \tab - \tab 10 \tab - \tab 1 \tab 3 \tab 12 \tab - \tab 2 \tab 2 \tab - \tab 5 \tab 21 \tab 5 \tab 2 \tab 3 \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 7 \tab 4 \tab 3 \tab 9 \tab - \tab 2 \tab - \tab 1 \tab 5 \tab 4 \tab 5 \tab 2 \tab 2 \tab 3 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab 12 \tab 13 \tab 6 \tab 9 \tab - \tab 8 \tab 1 \tab - \tab 6 \tab 751 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 7 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 12 \tab 1 \tab - \tab 1 \tab - \tab 7 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab 8 \tab - \tab 1 \tab 1 \tab 5 \tab 5 \tab 11 \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab 7 \tab 4 \tab 11 \tab 1 \tab 21 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 4 \tab 2 \tab 1 \tab - \tab - \tab 6 \tab 10 \tab 4 \tab - \tab - \tab 4 \tab 3 \tab 8 \tab 3 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab 8 \tab - \tab 1 \tab 1 \tab 8 \tab 10 \tab 1 \tab - \tab 5 \tab 8 \tab 1 \tab 11 \tab 10 \tab - \tab - \tab 1 \tab 1 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 6 \tab - \tab - \tab 2 \tab - \tab 5 \tab - \tab 2 \tab - \tab 14 \tab 2 \tab - \tab 1 \tab 1 \tab 15 \tab 21 \tab 3 \tab 1 \tab 31 \tab 1 \tab 5 \tab 1 \tab 3 \tab 3 \tab 3 \tab 4 \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 8 \tab - \tab 8 \tab - \tab 2 \tab - \tab 5 \tab 1 \tab - \tab 2 \tab 4 \tab 2 \tab 3 \tab - \tab 11 \tab 1 \tab - \tab 3 \tab 2 \tab 3 \tab 14 \tab 3 \tab 10 \tab 7 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 11 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 6 \tab 1 \tab 3 \tab - \tab 8 \tab 1 \tab 2 \tab 2 \tab 1 \tab 3 \tab 1 \tab 2 \tab 3 \tab 7 \tab 13 \tab 4 \tab 11 \tab 1 \tab 11 \tab 2 \tab - \tab 3 \tab 695 \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab 2 \tab 5 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 4 \tab - \tab - \tab - \tab 3 \tab 8 \tab 1 \tab - \tab 3 \tab 2 \tab 12 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab 7 \tab 3 \tab - \tab 5 \tab 1 \tab 5 \tab 3 \tab 9 \tab 4 \tab 19 \tab - \tab 5 \tab 1 \tab 3 \tab 3 \tab 6 \tab 8 \tab 3 \tab 20 \tab 1 \tab 7 \tab - \tab - \tab 3 \tab 2 \tab 7 \tab 1 \tab - \tab - \tab - \tab 12 \tab 3 \tab 4 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 9 \tab 8 \tab - \tab - \tab 1 \tab 12 \tab 4 \tab 4 \tab - \tab 1 \tab 2 \tab 2 \tab 5 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab - \tab 13 \tab 1 \tab 1 \tab 1 \tab 13 \tab 10 \tab - \tab - \tab 3 \tab 8 \tab - \tab 5 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab 4 \tab 2 \tab 5 \tab - \tab 1 \tab - \tab 20 \tab 1 \tab - \tab 2 \tab - \tab 21 \tab 22 \tab - \tab 2 \tab 36 \tab 3 \tab 2 \tab - \tab 2 \tab 2 \tab 4 \tab 2 \tab - \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 11 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab 6 \tab 3 \tab - \tab 3 \tab 3 \tab - \tab 2 \tab 3 \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 23 \tab 3 \tab 2 \tab 8 \tab - \tab 3 \tab - \tab - \tab - \tab 3 \tab 4 \tab 4 \tab 3 \tab 8 \tab 6 \tab - \tab 1 \tab - \tab 1 \tab 7 \tab 4 \tab - \tab 4 \tab 3 \tab 2 \tab - \tab 10 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab 2 \tab 17 \tab 17 \tab 13 \tab 7 \tab - \tab 5 \tab - \tab - \tab 5 \tab 757 \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 6 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab 4 \tab 19 \tab - \tab 1 \tab 3 \tab - \tab 5 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 7 \tab - \tab - \tab 6 \tab - \tab 4 \tab 6 \tab 3 \tab - \tab 13 \tab - \tab 3 \tab - \tab - \tab - \tab 7 \tab 10 \tab 9 \tab 17 \tab 1 \tab 11 \tab 1 \tab 1 \tab - \tab - \tab 10 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 2 \tab 5 \tab - \tab 2 \tab - \tab 4 \tab 8 \tab - \tab - \tab 1 \tab 8 \tab 5 \tab 3 \tab - \tab 3 \tab 1 \tab 3 \tab 4 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 9 \tab - \tab 1 \tab 1 \tab 7 \tab 6 \tab - \tab 2 \tab 8 \tab 8 \tab - \tab 12 \tab 6 \tab 2 \tab 2 \tab - \tab 3 \tab - \tab - \tab 5 \tab - \tab 1 \tab 3 \tab 3 \tab 1 \tab 2 \tab 7 \tab - \tab 3 \tab 2 \tab 1 \tab 9 \tab - \tab 2 \tab - \tab 22 \tab - \tab - \tab 2 \tab 1 \tab 20 \tab 19 \tab 5 \tab 1 \tab 40 \tab 3 \tab 6 \tab 1 \tab 4 \tab 2 \tab 2 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab 9 \tab 2 \tab 13 \tab - \tab 4 \tab 1 \tab 6 \tab 3 \tab 1 \tab 3 \tab 3 \tab - \tab 3 \tab 2 \tab 9 \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab 19 \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 5 \tab 6 \tab 5 \tab 3 \tab 4 \tab 3 \tab 16 \tab 8 \tab 2 \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab 4 \tab 9 \tab 4 \tab 2 \tab 1 \tab 9 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 4 \tab 19 \tab 13 \tab 6 \tab 5 \tab - \tab 11 \tab 1 \tab - \tab 4 \tab 792 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 4 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 6 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 3 \tab 2 \tab 2 \tab 3 \tab 2 \tab 6 \tab 2 \tab 2 \tab 2 \tab 15 \tab 1 \tab 1 \tab 3 \tab 2 \tab 14 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab 11 \tab 2 \tab - \tab 9 \tab 3 \tab 8 \tab 2 \tab 7 \tab 2 \tab 31 \tab - \tab 2 \tab - \tab - \tab 3 \tab 12 \tab 16 \tab 10 \tab 34 \tab 1 \tab 15 \tab 1 \tab - \tab 3 \tab 2 \tab 8 \tab 4 \tab 4 \tab 1 \tab 2 \tab 11 \tab 4 \tab 2 \tab 1 \tab 1 \tab 5 \tab - \tab 3 \tab - \tab 10 \tab 9 \tab 1 \tab - \tab 1 \tab 13 \tab 17 \tab 6 \tab - \tab 8 \tab 3 \tab 3 \tab 5 \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 22 \tab - \tab 2 \tab 1 \tab 17 \tab 18 \tab 1 \tab - \tab 8 \tab 15 \tab - \tab 14 \tab 15 \tab 2 \tab 2 \tab 2 \tab 3 \tab 2 \tab 4 \tab 7 \tab - \tab 1 \tab 4 \tab 3 \tab 2 \tab - \tab 7 \tab - \tab 3 \tab 8 \tab 1 \tab 13 \tab 1 \tab 4 \tab 2 \tab 41 \tab 4 \tab - \tab 5 \tab 5 \tab 34 \tab 43 \tab 8 \tab 1 \tab 68 \tab 1 \tab 9 \tab - \tab 9 \tab 11 \tab 6 \tab 4 \tab 5 \tab 5 \tab 3 \tab 2 \tab 2 \tab 2 \tab 21 \tab 2 \tab 9 \tab - \tab 1 \tab - \tab 12 \tab 7 \tab - \tab 7 \tab 6 \tab 2 \tab 5 \tab 2 \tab 34 \tab 3 \tab 2 \tab 1 \tab 4 \tab 10 \tab 27 \tab 4 \tab 12 \tab 5 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 6 \tab 6 \tab 4 \tab 8 \tab 23 \tab 16 \tab 1 \tab 4 \tab 1 \tab - \tab 4 \tab 9 \tab 8 \tab 17 \tab 8 \tab 3 \tab - \tab 12 \tab - \tab - \tab 12 \tab 1 \tab 4 \tab 2 \tab 6 \tab 4 \tab 31 \tab 27 \tab 19 \tab 17 \tab 3 \tab 10 \tab 3 \tab - \tab 6 \tab 1533 \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 3 \tab - \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab 2 \tab - \tab 1 \tab 13 \tab 6 \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab 5 \tab 3 \tab 1 \tab 5 \tab - \tab 3 \tab - \tab 4 \tab 2 \tab 11 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab 3 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab 4 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 13 \tab - \tab - \tab 5 \tab 21 \tab - \tab - \tab 1 \tab - \tab 11 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 11 \tab 4 \tab - \tab 8 \tab - \tab 3 \tab 3 \tab 9 \tab 1 \tab 32 \tab 1 \tab 2 \tab 1 \tab 3 \tab 3 \tab 9 \tab 21 \tab 7 \tab 36 \tab - \tab 8 \tab 1 \tab 1 \tab 4 \tab - \tab 9 \tab 1 \tab 4 \tab - \tab - \tab 10 \tab 4 \tab 3 \tab - \tab 2 \tab 8 \tab 1 \tab 2 \tab - \tab 11 \tab 6 \tab - \tab - \tab 1 \tab 11 \tab 11 \tab 6 \tab - \tab 8 \tab 3 \tab 6 \tab 8 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 13 \tab - \tab 1 \tab 2 \tab 16 \tab 24 \tab 1 \tab - \tab 9 \tab 12 \tab - \tab 4 \tab 13 \tab 1 \tab 4 \tab 1 \tab 4 \tab 1 \tab 4 \tab 6 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 8 \tab 1 \tab 2 \tab 1 \tab 4 \tab 13 \tab 1 \tab 4 \tab 1 \tab 36 \tab 4 \tab - \tab 3 \tab 6 \tab 35 \tab 28 \tab 7 \tab - \tab 51 \tab 4 \tab 12 \tab 2 \tab 5 \tab 8 \tab 10 \tab 1 \tab 4 \tab 5 \tab 4 \tab 2 \tab 9 \tab 2 \tab 14 \tab 2 \tab 7 \tab - \tab 4 \tab 2 \tab 7 \tab 8 \tab 1 \tab 6 \tab 3 \tab 2 \tab 2 \tab 1 \tab 13 \tab 3 \tab 2 \tab 4 \tab - \tab 9 \tab 24 \tab 5 \tab 9 \tab 10 \tab - \tab - \tab - \tab 6 \tab - \tab 3 \tab 4 \tab 5 \tab 5 \tab 18 \tab 3 \tab 3 \tab 1 \tab 4 \tab - \tab 6 \tab 3 \tab 8 \tab 17 \tab 7 \tab 3 \tab - \tab 24 \tab 1 \tab 1 \tab 16 \tab 5 \tab 4 \tab - \tab 3 \tab 2 \tab 23 \tab 29 \tab 16 \tab 8 \tab 2 \tab 14 \tab - \tab - \tab 16 \tab 1580 \tab - \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 3 \tab 2 \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab 2 \tab 4 \tab 2 \tab - \tab - \tab 2 \tab 3 \tab 3 \tab 1 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 9 \tab 4 \tab 4 \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 7 \tab - \tab 7 \tab 1 \tab 6 \tab 1 \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 3 \tab 4 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab 4 \tab 4 \tab - \tab 1 \tab 1 \tab 1 \tab 9 \tab 4 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 8 \tab 2 \tab - \tab 8 \tab 5 \tab 5 \tab 4 \tab 9 \tab 4 \tab 24 \tab - \tab 3 \tab - \tab 1 \tab 2 \tab 9 \tab 10 \tab 8 \tab 36 \tab 1 \tab 10 \tab 2 \tab 1 \tab 2 \tab - \tab 6 \tab 3 \tab 8 \tab 1 \tab - \tab 9 \tab 5 \tab 4 \tab - \tab - \tab 4 \tab 2 \tab 1 \tab 1 \tab 8 \tab 13 \tab - \tab 1 \tab 2 \tab 12 \tab 14 \tab 4 \tab - \tab 6 \tab 1 \tab 5 \tab 12 \tab 4 \tab - \tab - \tab 3 \tab - \tab - \tab 19 \tab - \tab 3 \tab 1 \tab 15 \tab 17 \tab 1 \tab 1 \tab 9 \tab 12 \tab 1 \tab 12 \tab 22 \tab 4 \tab 4 \tab - \tab 1 \tab 1 \tab 5 \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab 2 \tab 2 \tab 6 \tab - \tab 1 \tab 4 \tab 2 \tab 11 \tab 2 \tab 4 \tab 2 \tab 33 \tab 4 \tab - \tab 3 \tab 2 \tab 33 \tab 25 \tab 4 \tab - \tab 53 \tab 4 \tab 14 \tab 1 \tab 8 \tab 6 \tab 3 \tab 3 \tab - \tab 1 \tab 8 \tab - \tab - \tab - \tab 14 \tab 4 \tab 8 \tab - \tab 3 \tab 2 \tab 10 \tab 8 \tab 2 \tab 8 \tab - \tab 1 \tab 3 \tab 4 \tab 18 \tab 1 \tab 2 \tab 2 \tab 1 \tab 6 \tab 37 \tab 16 \tab 9 \tab 10 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 6 \tab 5 \tab 10 \tab 2 \tab 14 \tab 8 \tab 4 \tab 4 \tab 3 \tab - \tab 1 \tab 8 \tab 6 \tab 9 \tab 10 \tab 4 \tab - \tab 11 \tab - \tab - \tab 12 \tab 1 \tab 5 \tab 1 \tab 4 \tab 5 \tab 16 \tab 26 \tab 15 \tab 16 \tab 1 \tab 9 \tab 1 \tab - \tab 8 \tab 1549 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 5 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 4 \tab 2 \tab - \tab 11 \tab 3 \tab 4 \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab 6 \tab 3 \tab 1 \tab 4 \tab 1 \tab 13 \tab - \tab - \tab 1 \tab 9 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab 3 \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab 4 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab 1 \tab 7 \tab - \tab 3 \tab 1 \tab 5 \tab - \tab 1 \tab 3 \tab - \tab 9 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 3 \tab 4 \tab 2 \tab - \tab 9 \tab - \tab 3 \tab 3 \tab 9 \tab 5 \tab 22 \tab - \tab 5 \tab 1 \tab 1 \tab 3 \tab 16 \tab 12 \tab 15 \tab 37 \tab 1 \tab 12 \tab 2 \tab - \tab - \tab 1 \tab 8 \tab 1 \tab 4 \tab 4 \tab 2 \tab 6 \tab 2 \tab 4 \tab - \tab 1 \tab 5 \tab 2 \tab 2 \tab 1 \tab 8 \tab 12 \tab 1 \tab - \tab - \tab 13 \tab 11 \tab 5 \tab 1 \tab 5 \tab 2 \tab 1 \tab 5 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab - \tab 10 \tab - \tab 2 \tab 1 \tab 19 \tab 20 \tab - \tab - \tab 10 \tab 8 \tab - \tab 10 \tab 7 \tab - \tab 2 \tab 2 \tab - \tab - \tab 4 \tab 3 \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab - \tab 12 \tab - \tab - \tab 2 \tab 3 \tab 5 \tab - \tab 5 \tab - \tab 33 \tab - \tab 1 \tab 4 \tab 2 \tab 22 \tab 28 \tab 4 \tab 3 \tab 36 \tab 1 \tab 9 \tab 1 \tab 7 \tab 16 \tab 3 \tab 6 \tab 2 \tab 2 \tab 6 \tab 4 \tab 2 \tab 1 \tab 17 \tab 2 \tab 5 \tab 1 \tab 4 \tab 1 \tab 6 \tab 6 \tab - \tab 9 \tab 4 \tab - \tab 2 \tab - \tab 7 \tab 7 \tab 1 \tab 1 \tab 1 \tab 7 \tab 28 \tab 3 \tab 5 \tab 9 \tab 4 \tab - \tab - \tab 6 \tab 3 \tab 6 \tab 4 \tab 9 \tab 7 \tab 14 \tab 8 \tab 2 \tab 4 \tab 2 \tab 2 \tab 1 \tab 9 \tab 3 \tab 6 \tab 11 \tab 4 \tab 1 \tab 21 \tab 2 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 2 \tab - \tab 12 \tab 29 \tab 13 \tab 11 \tab 3 \tab 12 \tab 1 \tab - \tab 10 \tab 1648 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab - \tab 3 \tab 1 \tab - \tab 1 \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab 4 \tab 2 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 5 \tab 9 \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab 6 \tab - \tab - \tab 1 \tab 1 \tab 12 \tab - \tab 3 \tab - \tab 7 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 3 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 6 \tab - \tab 1 \tab 1 \tab 1 \tab 15 \tab 3 \tab - \tab - \tab 3 \tab - \tab 1 \tab 6 \tab 3 \tab 1 \tab 12 \tab - \tab 3 \tab 2 \tab 3 \tab 3 \tab 17 \tab 1 \tab 6 \tab - \tab 2 \tab 1 \tab 13 \tab 18 \tab 11 \tab 29 \tab 1 \tab 10 \tab - \tab - \tab 3 \tab - \tab 3 \tab 2 \tab 5 \tab 1 \tab - \tab 12 \tab 4 \tab 2 \tab - \tab 5 \tab 1 \tab 2 \tab - \tab 1 \tab 4 \tab 7 \tab - \tab - \tab 2 \tab 14 \tab 14 \tab 2 \tab 1 \tab 4 \tab 1 \tab 3 \tab 8 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab - \tab 21 \tab 1 \tab 5 \tab 1 \tab 13 \tab 15 \tab 3 \tab - \tab 10 \tab 12 \tab 1 \tab 10 \tab 18 \tab - \tab 4 \tab 4 \tab 3 \tab 1 \tab 7 \tab 4 \tab - \tab 1 \tab 3 \tab 5 \tab - \tab - \tab 5 \tab - \tab 2 \tab 3 \tab 3 \tab 8 \tab - \tab 5 \tab 1 \tab 25 \tab 1 \tab - \tab 4 \tab 5 \tab 28 \tab 28 \tab - \tab 1 \tab 50 \tab 3 \tab 6 \tab - \tab 9 \tab 13 \tab 2 \tab 2 \tab 1 \tab 2 \tab 4 \tab 2 \tab 3 \tab 2 \tab 21 \tab 2 \tab 9 \tab - \tab 4 \tab - \tab 9 \tab 8 \tab - \tab 7 \tab 4 \tab - \tab 3 \tab 5 \tab 11 \tab - \tab 4 \tab 4 \tab 2 \tab 2 \tab 29 \tab 6 \tab 6 \tab 11 \tab 4 \tab 1 \tab 1 \tab 3 \tab 5 \tab 6 \tab 9 \tab 7 \tab 7 \tab 11 \tab 5 \tab 3 \tab 2 \tab 4 \tab - \tab 1 \tab 5 \tab 3 \tab 14 \tab 3 \tab 5 \tab - \tab 16 \tab 1 \tab 1 \tab 3 \tab 1 \tab 6 \tab 3 \tab 2 \tab 4 \tab 12 \tab 27 \tab 21 \tab 11 \tab 1 \tab 12 \tab - \tab 1 \tab 13 \tab 1565 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 5 \tab 1 \tab 2 \tab 3 \tab 4 \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab 4 \tab 5 \tab 2 \tab 1 \tab 5 \tab 1 \tab 11 \tab - \tab 2 \tab 2 \tab 9 \tab - \tab 3 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 3 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 7 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1 \tab 1 \tab - \tab 3 \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab 8 \tab - \tab - \tab 4 \tab 1 \tab 16 \tab 1 \tab - \tab - \tab - \tab 2 \tab 3 \tab 18 \tab 1 \tab - \tab 10 \tab - \tab 6 \tab 4 \tab 5 \tab 3 \tab 25 \tab 1 \tab 2 \tab - \tab - \tab - \tab 8 \tab 15 \tab 6 \tab 38 \tab - \tab 16 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab - \tab 5 \tab 1 \tab - \tab 11 \tab 6 \tab 2 \tab 1 \tab 1 \tab 5 \tab 4 \tab 1 \tab - \tab 7 \tab 3 \tab - \tab 1 \tab 1 \tab 12 \tab 8 \tab 4 \tab 1 \tab 8 \tab - \tab 2 \tab 13 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 19 \tab - \tab - \tab 1 \tab 17 \tab 20 \tab 1 \tab 1 \tab 12 \tab 5 \tab - \tab 9 \tab 15 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 7 \tab 6 \tab - \tab - \tab 4 \tab 5 \tab - \tab - \tab 13 \tab 1 \tab 2 \tab 5 \tab 2 \tab 8 \tab - \tab 5 \tab 1 \tab 41 \tab 4 \tab - \tab 7 \tab 6 \tab 33 \tab 24 \tab 2 \tab 2 \tab 46 \tab 3 \tab 17 \tab - \tab 2 \tab 12 \tab 5 \tab 1 \tab 3 \tab 2 \tab 10 \tab 1 \tab 3 \tab - \tab 21 \tab 1 \tab 11 \tab - \tab 4 \tab - \tab 5 \tab 6 \tab - \tab 3 \tab 3 \tab 2 \tab 4 \tab 3 \tab 14 \tab 4 \tab 4 \tab 1 \tab 3 \tab 5 \tab 35 \tab 7 \tab 13 \tab 12 \tab 1 \tab - \tab 1 \tab 3 \tab 4 \tab 5 \tab 8 \tab 13 \tab 6 \tab 9 \tab 11 \tab 2 \tab 2 \tab 1 \tab 2 \tab 4 \tab 9 \tab 3 \tab 16 \tab 5 \tab 4 \tab 1 \tab 27 \tab 1 \tab - \tab 11 \tab 4 \tab 4 \tab 1 \tab 2 \tab 1 \tab 22 \tab 41 \tab 19 \tab 18 \tab 4 \tab 15 \tab 1 \tab - \tab 15 \tab 1448 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 2 \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1 \tab 5 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 2 \tab 3 \tab 2 \tab 4 \tab 2 \tab 12 \tab 7 \tab 6 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 6 \tab 3 \tab 5 \tab 1 \tab 5 \tab 3 \tab 5 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 8 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab 4 \tab 13 \tab - \tab 2 \tab 2 \tab 19 \tab - \tab - \tab 1 \tab 2 \tab 21 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 3 \tab 19 \tab 3 \tab 1 \tab 25 \tab 4 \tab 10 \tab 10 \tab 15 \tab 3 \tab 53 \tab 1 \tab 9 \tab - \tab - \tab - \tab 12 \tab 30 \tab 13 \tab 64 \tab 2 \tab 26 \tab 4 \tab - \tab 4 \tab - \tab 15 \tab 3 \tab 3 \tab 2 \tab - \tab 17 \tab 9 \tab 11 \tab 1 \tab - \tab 8 \tab 3 \tab - \tab - \tab 12 \tab 13 \tab - \tab - \tab 2 \tab 24 \tab 15 \tab 7 \tab - \tab 8 \tab - \tab 7 \tab 12 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab 34 \tab - \tab 4 \tab 2 \tab 23 \tab 26 \tab - \tab 1 \tab 13 \tab 9 \tab - \tab 10 \tab 18 \tab 5 \tab 6 \tab 6 \tab 6 \tab - \tab 8 \tab 6 \tab 1 \tab 1 \tab 3 \tab 4 \tab 1 \tab 1 \tab 14 \tab - \tab 2 \tab 9 \tab 7 \tab 4 \tab 1 \tab 10 \tab 3 \tab 41 \tab 7 \tab - \tab 5 \tab 4 \tab 27 \tab 37 \tab 2 \tab 4 \tab 63 \tab 8 \tab 18 \tab 1 \tab 12 \tab 20 \tab 5 \tab 2 \tab 5 \tab 3 \tab 5 \tab 3 \tab 5 \tab 1 \tab 14 \tab 4 \tab 14 \tab - \tab 4 \tab - \tab 15 \tab 5 \tab 1 \tab 5 \tab 12 \tab 3 \tab 6 \tab 3 \tab 21 \tab 3 \tab 3 \tab 5 \tab - \tab 9 \tab 42 \tab 14 \tab 12 \tab 14 \tab - \tab 1 \tab 1 \tab 3 \tab 6 \tab 4 \tab 10 \tab 16 \tab 4 \tab 20 \tab 7 \tab 5 \tab 3 \tab 4 \tab 1 \tab 5 \tab 11 \tab 11 \tab 20 \tab 20 \tab 7 \tab 2 \tab 30 \tab - \tab 1 \tab 9 \tab 7 \tab 7 \tab 1 \tab 3 \tab 5 \tab 36 \tab 63 \tab 23 \tab 24 \tab 5 \tab 13 \tab 1 \tab 1 \tab 22 \tab 2508 \tab - \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab 4 \tab 9 \tab - \tab 4 \tab - \tab 3 \tab 5 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab - \tab 5 \tab 20 \tab 6 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab 4 \tab 6 \tab 4 \tab - \tab 11 \tab 2 \tab 10 \tab - \tab 6 \tab - \tab 22 \tab 2 \tab 6 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab 3 \tab 1 \tab 6 \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 4 \tab 3 \tab 2 \tab 7 \tab 4 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 9 \tab 2 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 4 \tab - \tab 1 \tab 3 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 5 \tab 1 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab 3 \tab - \tab 11 \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 8 \tab - \tab - \tab 8 \tab 2 \tab 8 \tab 2 \tab 9 \tab 3 \tab 20 \tab - \tab 3 \tab 1 \tab - \tab 3 \tab 10 \tab 11 \tab 7 \tab 31 \tab - \tab 7 \tab 1 \tab 1 \tab 3 \tab - \tab 6 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab 7 \tab 8 \tab 3 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab - \tab 16 \tab 1 \tab 1 \tab - \tab 7 \tab 15 \tab 1 \tab - \tab 8 \tab 7 \tab - \tab 8 \tab 9 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab 8 \tab 1 \tab 2 \tab 5 \tab - \tab 7 \tab - \tab 6 \tab - \tab 14 \tab 7 \tab - \tab 1 \tab - \tab 21 \tab 18 \tab 1 \tab - \tab 41 \tab 3 \tab 10 \tab 1 \tab 11 \tab 11 \tab 5 \tab - \tab 2 \tab 3 \tab 4 \tab 2 \tab - \tab - \tab 7 \tab - \tab 5 \tab - \tab 5 \tab 2 \tab 4 \tab 3 \tab - \tab 3 \tab 5 \tab 2 \tab 2 \tab 3 \tab 8 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 21 \tab 8 \tab 3 \tab 5 \tab 1 \tab - \tab - \tab 7 \tab 2 \tab 4 \tab 6 \tab 12 \tab 5 \tab 13 \tab 5 \tab 2 \tab 1 \tab 3 \tab - \tab 3 \tab 7 \tab 3 \tab 9 \tab 2 \tab - \tab - \tab 11 \tab 2 \tab 2 \tab 5 \tab 3 \tab 4 \tab - \tab 2 \tab 2 \tab 16 \tab 14 \tab 15 \tab 14 \tab 1 \tab 9 \tab 1 \tab - \tab 3 \tab 1095 \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 5 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 5 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 6 \tab 1 \tab 12 \tab - \tab 3 \tab 2 \tab 10 \tab 1 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 7 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 4 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 4 \tab - \tab - \tab 2 \tab 8 \tab - \tab 1 \tab 3 \tab - \tab 7 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 10 \tab 3 \tab - \tab 8 \tab 3 \tab 9 \tab 2 \tab 5 \tab 1 \tab 24 \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 5 \tab 4 \tab 3 \tab 26 \tab - \tab 6 \tab 1 \tab 2 \tab 1 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab 1 \tab 5 \tab 1 \tab 2 \tab 3 \tab 1 \tab 2 \tab - \tab - \tab - \tab 6 \tab 6 \tab - \tab - \tab 2 \tab 12 \tab 4 \tab 1 \tab - \tab 3 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 15 \tab - \tab - \tab - \tab 7 \tab 10 \tab 1 \tab - \tab 6 \tab 3 \tab - \tab 3 \tab 8 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab 5 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 2 \tab 4 \tab 3 \tab 5 \tab 1 \tab 2 \tab 1 \tab 11 \tab 2 \tab - \tab 1 \tab 4 \tab 16 \tab 18 \tab - \tab 3 \tab 25 \tab 3 \tab 9 \tab 1 \tab 3 \tab 7 \tab 1 \tab 1 \tab 5 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab 15 \tab 3 \tab 7 \tab - \tab 3 \tab 1 \tab 4 \tab 2 \tab - \tab 3 \tab 2 \tab - \tab 2 \tab 2 \tab 9 \tab 1 \tab 2 \tab 1 \tab 1 \tab 4 \tab 19 \tab 4 \tab 8 \tab 4 \tab - \tab - \tab - \tab 4 \tab 1 \tab 7 \tab 3 \tab 12 \tab 4 \tab 7 \tab 3 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab 6 \tab 3 \tab 14 \tab 11 \tab - \tab 1 \tab 9 \tab - \tab - \tab 5 \tab - \tab 1 \tab - \tab 2 \tab 4 \tab 12 \tab 15 \tab 13 \tab 7 \tab 2 \tab 8 \tab 1 \tab - \tab 11 \tab 1205 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 6 \tab - \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab 3 \tab 2 \tab 2 \tab 8 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 4 \tab 3 \tab 2 \tab 5 \tab 2 \tab 5 \tab 1 \tab 1 \tab 2 \tab 6 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab 7 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab 4 \tab 1 \tab - \tab 4 \tab 1 \tab 11 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab 12 \tab - \tab 6 \tab 3 \tab 6 \tab - \tab 23 \tab - \tab - \tab 1 \tab 1 \tab - \tab 6 \tab 10 \tab 6 \tab 27 \tab - \tab 5 \tab - \tab 1 \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab 7 \tab 1 \tab 3 \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab 3 \tab 3 \tab - \tab - \tab - \tab 6 \tab 5 \tab 4 \tab - \tab 7 \tab 1 \tab 3 \tab 10 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab 4 \tab 8 \tab 1 \tab - \tab 6 \tab 2 \tab - \tab 3 \tab 6 \tab 1 \tab 1 \tab 5 \tab 1 \tab - \tab 7 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 4 \tab - \tab 1 \tab 2 \tab 1 \tab 9 \tab 2 \tab 3 \tab - \tab 11 \tab - \tab 1 \tab 1 \tab 1 \tab 17 \tab 16 \tab - \tab - \tab 26 \tab 2 \tab 15 \tab 1 \tab 6 \tab 8 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 5 \tab 3 \tab 4 \tab - \tab 1 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 9 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5 \tab 41 \tab 3 \tab 8 \tab 5 \tab 1 \tab - \tab - \tab 1 \tab 5 \tab 3 \tab 4 \tab 10 \tab 1 \tab 12 \tab 9 \tab 2 \tab - \tab 4 \tab 2 \tab 3 \tab 3 \tab 6 \tab 8 \tab 6 \tab 10 \tab - \tab 12 \tab 1 \tab 2 \tab 1 \tab 2 \tab 5 \tab - \tab 1 \tab - \tab 8 \tab 16 \tab 8 \tab 15 \tab 3 \tab 11 \tab - \tab - \tab 14 \tab 1149 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 5 \tab - \tab 4 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 4 \tab 18 \tab 1 \tab 6 \tab - \tab 10 \tab 4 \tab 1 \tab - \tab 7 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab 5 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 2 \tab 4 \tab - \tab 1 \tab 1 \tab 10 \tab - \tab - \tab 2 \tab - \tab 11 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 7 \tab - \tab - \tab 16 \tab - \tab 9 \tab 2 \tab 6 \tab - \tab 28 \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 14 \tab 11 \tab 8 \tab 30 \tab - \tab 9 \tab - \tab 1 \tab 5 \tab - \tab 5 \tab 3 \tab 1 \tab - \tab 2 \tab 7 \tab 9 \tab 4 \tab 1 \tab 2 \tab 3 \tab 1 \tab - \tab - \tab 5 \tab 4 \tab 2 \tab - \tab - \tab 4 \tab 3 \tab 7 \tab - \tab 5 \tab - \tab 5 \tab 5 \tab 2 \tab 2 \tab - \tab 2 \tab - \tab - \tab 17 \tab - \tab - \tab - \tab 8 \tab 5 \tab - \tab - \tab 10 \tab 8 \tab - \tab 1 \tab 6 \tab 1 \tab 4 \tab - \tab 2 \tab - \tab 4 \tab 3 \tab - \tab - \tab 2 \tab 5 \tab - \tab - \tab 5 \tab - \tab 1 \tab 3 \tab - \tab 8 \tab - \tab 3 \tab - \tab 21 \tab 4 \tab - \tab 2 \tab 2 \tab 20 \tab 21 \tab - \tab - \tab 31 \tab 4 \tab 14 \tab 1 \tab 6 \tab 10 \tab 2 \tab - \tab 1 \tab 2 \tab 4 \tab - \tab 1 \tab 4 \tab 12 \tab 5 \tab 7 \tab - \tab 1 \tab 1 \tab 10 \tab 5 \tab - \tab 4 \tab 13 \tab 2 \tab 7 \tab 2 \tab 15 \tab 3 \tab 2 \tab 3 \tab 1 \tab 6 \tab 20 \tab 6 \tab 9 \tab 9 \tab 2 \tab 1 \tab - \tab 4 \tab 3 \tab 6 \tab 5 \tab 14 \tab 8 \tab 10 \tab 5 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 13 \tab 2 \tab 12 \tab 8 \tab 3 \tab - \tab 17 \tab - \tab 1 \tab 6 \tab 4 \tab 3 \tab - \tab 2 \tab 4 \tab 16 \tab 17 \tab 20 \tab 14 \tab 3 \tab 12 \tab 3 \tab - \tab 15 \tab 1455 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 6 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 4 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 6 \tab 3 \tab 2 \tab 2 \tab 2 \tab - \tab - \tab 3 \tab 8 \tab 2 \tab - \tab 3 \tab - \tab 10 \tab - \tab 3 \tab 1 \tab 15 \tab 1 \tab 5 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab - \tab 4 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 3 \tab 2 \tab 1 \tab 10 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 6 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 3 \tab 2 \tab 9 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 13 \tab 2 \tab - \tab 13 \tab 5 \tab 5 \tab 4 \tab 12 \tab 3 \tab 34 \tab 1 \tab 1 \tab - \tab - \tab - \tab 17 \tab 8 \tab 3 \tab 20 \tab - \tab 13 \tab 1 \tab 1 \tab 3 \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab 11 \tab 5 \tab 3 \tab 1 \tab 3 \tab 5 \tab - \tab - \tab 1 \tab 5 \tab 8 \tab - \tab - \tab 2 \tab 16 \tab 5 \tab 4 \tab 1 \tab 7 \tab 1 \tab 1 \tab 8 \tab - \tab 2 \tab - \tab 3 \tab - \tab - \tab 11 \tab - \tab - \tab 2 \tab 5 \tab 11 \tab - \tab - \tab 9 \tab 6 \tab 1 \tab 5 \tab 5 \tab 3 \tab 3 \tab - \tab 3 \tab 1 \tab 1 \tab 5 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 8 \tab - \tab 3 \tab 5 \tab 3 \tab 12 \tab - \tab 5 \tab 1 \tab 22 \tab 4 \tab - \tab - \tab 3 \tab 19 \tab 16 \tab - \tab 4 \tab 23 \tab 5 \tab 13 \tab 1 \tab 11 \tab 5 \tab 3 \tab 4 \tab 1 \tab 1 \tab 4 \tab - \tab - \tab - \tab 13 \tab 2 \tab 3 \tab - \tab 4 \tab 3 \tab 11 \tab 7 \tab - \tab 3 \tab 5 \tab - \tab - \tab 3 \tab 16 \tab 1 \tab 2 \tab - \tab 1 \tab 5 \tab 34 \tab 3 \tab 11 \tab 11 \tab - \tab - \tab - \tab 4 \tab 2 \tab 8 \tab 10 \tab 5 \tab 9 \tab 12 \tab 5 \tab 6 \tab 1 \tab 3 \tab - \tab 1 \tab 6 \tab 2 \tab 13 \tab 5 \tab 3 \tab - \tab 19 \tab 1 \tab 1 \tab 12 \tab 9 \tab 4 \tab 2 \tab 1 \tab 1 \tab 27 \tab 29 \tab 23 \tab 19 \tab 1 \tab 22 \tab 1 \tab - \tab 7 \tab 1723 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 8 \tab 2 \tab 3 \tab - \tab 1 \tab 4 \tab - \tab - \tab 3 \tab 1 \tab 1 \tab 2 \tab - \tab 3 \tab 1 \tab 1 \tab 12 \tab 3 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 10 \tab 1 \tab - \tab 4 \tab - \tab 15 \tab - \tab 2 \tab 1 \tab 10 \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 3 \tab 4 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 6 \tab - \tab 1 \tab 3 \tab 7 \tab - \tab 2 \tab 2 \tab 1 \tab 10 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab 9 \tab 4 \tab 1 \tab 15 \tab 1 \tab 9 \tab 3 \tab 10 \tab 3 \tab 39 \tab - \tab 4 \tab - \tab - \tab 1 \tab 8 \tab 9 \tab 7 \tab 39 \tab 2 \tab 11 \tab 2 \tab 1 \tab 1 \tab - \tab 6 \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 3 \tab - \tab 2 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab 3 \tab 8 \tab 2 \tab - \tab - \tab 11 \tab 8 \tab 3 \tab - \tab 6 \tab 1 \tab 4 \tab 9 \tab 1 \tab 2 \tab 1 \tab 4 \tab - \tab - \tab 22 \tab 1 \tab - \tab - \tab 4 \tab 17 \tab 2 \tab - \tab 12 \tab 2 \tab 2 \tab 6 \tab 11 \tab 2 \tab 2 \tab 1 \tab - \tab 1 \tab 4 \tab 5 \tab - \tab - \tab 5 \tab 3 \tab 1 \tab - \tab 5 \tab 1 \tab 4 \tab 3 \tab - \tab 3 \tab 1 \tab 5 \tab - \tab 16 \tab 2 \tab 2 \tab - \tab 3 \tab 18 \tab 14 \tab 1 \tab 3 \tab 24 \tab 6 \tab 12 \tab - \tab 5 \tab 4 \tab 6 \tab 4 \tab 4 \tab 6 \tab 9 \tab - \tab - \tab - \tab 14 \tab 4 \tab 5 \tab - \tab 3 \tab - \tab 5 \tab 6 \tab 1 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab 8 \tab 4 \tab 5 \tab 3 \tab 1 \tab 7 \tab 28 \tab 6 \tab 11 \tab 8 \tab - \tab - \tab - \tab 3 \tab 6 \tab 9 \tab 12 \tab 10 \tab 13 \tab 10 \tab 11 \tab 4 \tab 1 \tab 2 \tab 1 \tab 1 \tab 7 \tab 7 \tab 8 \tab 7 \tab 7 \tab - \tab 15 \tab - \tab 1 \tab 6 \tab 4 \tab 3 \tab 3 \tab 4 \tab - \tab 23 \tab 21 \tab 24 \tab 15 \tab 5 \tab 12 \tab - \tab 1 \tab 9 \tab 1179 \tab 2 \tab - \tab 2 \tab 1 \tab 2 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 6 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 2 \tab 7 \tab 1 \tab 1 \tab 6 \tab - \tab 11 \tab 1 \tab 5 \tab 1 \tab 14 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 2 \tab - \tab 1 \tab 7 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab - \tab 3 \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 2 \tab - \tab 3 \tab 8 \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 3 \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 24 \tab 49 \tab 28 \tab 36 \tab 39 \tab 55 \tab 187 \tab 8 \tab 29 \tab 76 \tab 429 \tab 8 \tab 25 \tab 92 \tab 36 \tab 353 \tab 45 \tab 12 \tab 24 \tab 29 \tab 15 \tab 46 \tab 272 \tab 71 \tab 5 \tab 256 \tab 44 \tab 138 \tab 98 \tab 266 \tab 92 \tab 749 \tab 15 \tab 99 \tab 48 \tab 15 \tab 32 \tab 296 \tab 355 \tab 234 \tab 847 \tab 24 \tab 385 \tab 36 \tab 25 \tab 62 \tab 20 \tab 173 \tab 53 \tab 79 \tab 32 \tab 16 \tab 255 \tab 109 \tab 84 \tab 40 \tab 43 \tab 101 \tab 42 \tab 22 \tab 11 \tab 205 \tab 227 \tab 14 \tab 6 \tab 39 \tab 326 \tab 254 \tab 117 \tab 11 \tab 133 \tab 51 \tab 110 \tab 223 \tab 47 \tab 60 \tab 10 \tab 48 \tab 6 \tab 13 \tab 444 \tab 8 \tab 43 \tab 33 \tab 395 \tab 422 \tab 31 \tab 11 \tab 212 \tab 300 \tab 11 \tab 389 \tab 274 \tab 49 \tab 75 \tab 42 \tab 77 \tab 13 \tab 114 \tab 108 \tab 20 \tab 21 \tab 61 \tab 101 \tab 20 \tab 56 \tab 193 \tab 21 \tab 64 \tab 127 \tab 57 \tab 218 \tab 22 \tab 76 \tab 19 \tab 725 \tab 76 \tab 5 \tab 100 \tab 90 \tab 705 \tab 779 \tab 109 \tab 55 \tab 1327 \tab 74 \tab 228 \tab 14 \tab 153 \tab 160 \tab 106 \tab 79 \tab 51 \tab 65 \tab 129 \tab 37 \tab 50 \tab 31 \tab 369 \tab 41 \tab 214 \tab 8 \tab 71 \tab 26 \tab 170 \tab 148 \tab 12 \tab 117 \tab 162 \tab 26 \tab 69 \tab 63 \tab 335 \tab 54 \tab 49 \tab 58 \tab 30 \tab 172 \tab 646 \tab 159 \tab 197 \tab 175 \tab 59 \tab 22 \tab 9 \tab 85 \tab 64 \tab 112 \tab 132 \tab 214 \tab 125 \tab 284 \tab 172 \tab 50 \tab 44 \tab 57 \tab 18 \tab 80 \tab 134 \tab 102 \tab 260 \tab 145 \tab 91 \tab 15 \tab 371 \tab 21 \tab 24 \tab 158 \tab 67 \tab 94 \tab 21 \tab 53 \tab 71 \tab 493 \tab 696 \tab 433 \tab 314 \tab 40 \tab 261 \tab 23 \tab 5 \tab 234 \tab 37061 \tab 11 \tab 21 \tab 86 \tab 12 \tab 31 \tab 14 \tab 21 \tab 22 \tab 17 \tab 10 \tab 33 \tab 16 \tab 44 \tab 14 \tab 11 \tab 70 \tab 7 \tab 7 \tab 5 \tab 2 \tab 33 \tab 87 \tab 28 \tab 38 \tab 12 \tab 36 \tab 64 \tab 31 \tab 10 \tab 24 \tab 20 \tab 17 \tab 33 \tab 50 \tab 51 \tab 30 \tab 35 \tab 167 \tab 51 \tab 38 \tab 17 \tab 16 \tab 16 \tab 24 \tab 60 \tab 106 \tab 60 \tab 24 \tab 120 \tab 16 \tab 175 \tab 17 \tab 59 \tab 28 \tab 162 \tab 14 \tab 38 \tab 22 \tab 3 \tab 11 \tab 28 \tab 10 \tab 14 \tab 7 \tab 25 \tab 46 \tab 20 \tab 49 \tab 28 \tab 20 \tab 50 \tab 25 \tab 12 \tab 11 \tab 32 \tab 5 \tab 23 \tab 8 \tab 12 \tab 48 \tab 46 \tab 38 \tab 40 \tab 14 \tab 10 \tab 12 \tab 2 \tab 16 \tab 14 \tab 20 \tab 7 \tab 27 \tab 10 \tab 46 \tab 79 \tab 24 \tab 8 \tab 5 \tab 28 \tab 8 \tab 14 \tab 9 \tab 3 \tab 19 \tab 3 \tab 20 \tab 26 \tab 1 \tab 20 \tab 14 \tab 4 \tab 18 \tab 12 \tab 7 \tab 13 \tab 21 \tab 26 \tab 24 \tab 3 \tab 15 \tab 19 \tab 11 \tab 4 \tab 6 \tab 5 \tab 5 \tab 29 \tab 10 \tab 3 \tab 20 \tab 8 \tab 12 \tab 25 \tab 18 \tab 6 \tab 8 \tab 5 \tab 4 \tab 4 \tab 14 \tab 7 \tab 2 \tab 2 \tab 5 \tab 19 \tab 9 \tab 7 \tab 2 \tab 5 \tab 12 \tab 8 \tab 10 \tab 1 \tab 4 \tab 10 \tab 9 \tab 16 \tab 6 \tab 9 \tab 2 \tab 7 \tab 6 \tab 7 \tab 2 \tab 10 \tab 5 \tab 5 \tab 2 \tab 1 \tab 4 \tab 1 \tab 2 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} i don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
