% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_17.R
\name{srcbelt}
\alias{srcbelt}
\title{srcbelt}
\source{
General Social Survey https://gss.norc.org
}
\description{
Src beltcode
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab central city of 12 largest smsas \tab central city of the remainder of the 100 largest smsas \tab other rural (counties having no towns of 10,000 or more) \tab other urban (counties having towns of 10,000 or more) \tab suburbs of 12 largest smsas \tab suburbs of the remaining 100 largest smsas \tab iap \tab Total \cr
   1972 \tab 213 \tab 299 \tab 267 \tab 473 \tab 218 \tab 143 \tab - \tab 1613 \cr
   1973 \tab 173 \tab 236 \tab 259 \tab 508 \tab 134 \tab 194 \tab - \tab 1504 \cr
   1974 \tab 169 \tab 223 \tab 262 \tab 508 \tab 132 \tab 190 \tab - \tab 1484 \cr
   1975 \tab 157 \tab 225 \tab 255 \tab 579 \tab 116 \tab 158 \tab - \tab 1490 \cr
   1976 \tab 160 \tab 186 \tab 290 \tab 562 \tab 134 \tab 167 \tab - \tab 1499 \cr
   1977 \tab 134 \tab 214 \tab 273 \tab 608 \tab 111 \tab 190 \tab - \tab 1530 \cr
   1978 \tab 106 \tab 228 \tab 292 \tab 551 \tab 181 \tab 174 \tab - \tab 1532 \cr
   1980 \tab 123 \tab 219 \tab 253 \tab 578 \tab 129 \tab 166 \tab - \tab 1468 \cr
   1982 \tab 230 \tab 298 \tab 308 \tab 717 \tab 150 \tab 157 \tab - \tab 1860 \cr
   1983 \tab 124 \tab 160 \tab 213 \tab 686 \tab 206 \tab 210 \tab - \tab 1599 \cr
   1984 \tab 128 \tab 211 \tab 185 \tab 499 \tab 165 \tab 285 \tab - \tab 1473 \cr
   1985 \tab 127 \tab 142 \tab 274 \tab 554 \tab 162 \tab 275 \tab - \tab 1534 \cr
   1986 \tab 117 \tab 186 \tab 217 \tab 486 \tab 168 \tab 296 \tab - \tab 1470 \cr
   1987 \tab 201 \tab 271 \tab 201 \tab 646 \tab 164 \tab 336 \tab - \tab 1819 \cr
   1988 \tab 115 \tab 170 \tab 166 \tab 547 \tab 174 \tab 309 \tab - \tab 1481 \cr
   1989 \tab 107 \tab 210 \tab 214 \tab 599 \tab 165 \tab 242 \tab - \tab 1537 \cr
   1990 \tab 106 \tab 170 \tab 196 \tab 496 \tab 165 \tab 239 \tab - \tab 1372 \cr
   1991 \tab 131 \tab 185 \tab 197 \tab 610 \tab 170 \tab 224 \tab - \tab 1517 \cr
   1993 \tab 110 \tab 237 \tab 183 \tab 609 \tab 220 \tab 247 \tab - \tab 1606 \cr
   1994 \tab 257 \tab 509 \tab 264 \tab 1181 \tab 333 \tab 448 \tab - \tab 2992 \cr
   1996 \tab 233 \tab 483 \tab 279 \tab 1189 \tab 309 \tab 411 \tab - \tab 2904 \cr
   1998 \tab 249 \tab 475 \tab 283 \tab 1110 \tab 305 \tab 410 \tab - \tab 2832 \cr
   2000 \tab 235 \tab 449 \tab 296 \tab 1152 \tab 334 \tab 351 \tab - \tab 2817 \cr
   2002 \tab 213 \tab 411 \tab 286 \tab 1146 \tab 333 \tab 376 \tab - \tab 2765 \cr
   2004 \tab 197 \tab 284 \tab 362 \tab 1180 \tab 414 \tab 375 \tab - \tab 2812 \cr
   2006 \tab 374 \tab 463 \tab 570 \tab 1804 \tab 593 \tab 706 \tab - \tab 4510 \cr
   2008 \tab 176 \tab 221 \tab 249 \tab 805 \tab 217 \tab 355 \tab - \tab 2023 \cr
   2010 \tab 180 \tab 258 \tab 226 \tab 819 \tab 226 \tab 335 \tab - \tab 2044 \cr
   2012 \tab 187 \tab 258 \tab 257 \tab 650 \tab 256 \tab 366 \tab - \tab 1974 \cr
   2014 \tab 279 \tab 357 \tab 291 \tab 780 \tab 358 \tab 473 \tab - \tab 2538 \cr
   2016 \tab 261 \tab 432 \tab 312 \tab 951 \tab 371 \tab 540 \tab - \tab 2867 \cr
   2018 \tab 233 \tab 345 \tab 259 \tab 741 \tab 332 \tab 438 \tab - \tab 2348 \cr
   2021 \tab 333 \tab 650 \tab 419 \tab 1293 \tab 553 \tab 784 \tab - \tab 4032 \cr
   2022 \tab 420 \tab 541 \tab 356 \tab 1023 \tab 512 \tab 634 \tab 58 \tab 3544 \cr
   Total \tab 6558 \tab 10206 \tab 9214 \tab 26640 \tab 8510 \tab 11204 \tab 58 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} central city of 12 largest smsas
\item \code{2} central city of the remainder of the 100 largest smsas
\item \code{3} suburbs of 12 largest smsas
\item \code{4} suburbs of the remaining 100 largest smsas
\item \code{5} other urban (counties having towns of 10,000 or more)
\item \code{6} other rural (counties having no towns of 10,000 or more)
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
