% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_500.R
\name{stress}
\alias{stress}
\title{stress}
\source{
General Social Survey https://gss.norc.org
}
\description{
How often does r find work stressful
}
\details{
Question 1336. Now some more questions about your working conditions. Please circle one code for each item below to show how often it applies to your work. How often... c. Do you find your work stressful?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not available in this year \tab always \tab can't choose \tab hardly ever \tab never \tab no answer \tab often \tab sometimes \tab skipped on web \tab Total \cr
   1972 \tab 3 \tab 1610 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab - \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab - \tab 82 \tab 3 \tab 109 \tab 37 \tab 663 \tab 261 \tab 382 \tab - \tab 1537 \cr
   1990 \tab - \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab - \tab 89 \tab 1 \tab 91 \tab 42 \tab 2002 \tab 229 \tab 378 \tab - \tab 2832 \cr
   2000 \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab - \tab 161 \tab 6 \tab 329 \tab 137 \tab 985 \tab 391 \tab 756 \tab - \tab 2765 \cr
   2004 \tab - \tab - \tab 213 \tab 5 \tab 242 \tab 95 \tab 985 \tab 475 \tab 797 \tab - \tab 2812 \cr
   2006 \tab - \tab - \tab 353 \tab 3 \tab 449 \tab 185 \tab 1771 \tab 595 \tab 1154 \tab - \tab 4510 \cr
   2008 \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab 110 \tab 1 \tab 215 \tab 78 \tab 881 \tab 284 \tab 475 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab 118 \tab 5 \tab 162 \tab 66 \tab 825 \tab 320 \tab 478 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab 113 \tab 1 \tab 225 \tab 99 \tab 1294 \tab 247 \tab 559 \tab - \tab 2538 \cr
   2016 \tab 911 \tab - \tab 96 \tab 2 \tab 130 \tab 32 \tab 1022 \tab 207 \tab 467 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab 107 \tab 3 \tab 264 \tab 110 \tab 932 \tab 304 \tab 628 \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 1557 \tab - \tab 171 \tab 5 \tab 339 \tab 128 \tab 14 \tab 409 \tab 906 \tab 15 \tab 3544 \cr
   Total \tab 6503 \tab 38584 \tab 1613 \tab 35 \tab 2555 \tab 1009 \tab 11374 \tab 3722 \tab 6980 \tab 15 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} hardly ever
\item \code{5} never
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
