% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_72.R
\name{suicide1}
\alias{suicide1}
\title{suicide1}
\source{
General Social Survey https://gss.norc.org
}
\description{
Suicide if incurable disease
}
\details{
Question 227. Do you think a person has the right to end his or her own life if this person: a. Has an incurable disease?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 49 \tab 897 \tab 6 \tab 578 \tab - \tab 1530 \cr
   1978 \tab - \tab 49 \tab 895 \tab - \tab 588 \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 82 \tab 1014 \tab 10 \tab 754 \tab - \tab 1860 \cr
   1983 \tab - \tab 61 \tab 770 \tab 8 \tab 760 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 52 \tab 805 \tab 2 \tab 675 \tab - \tab 1534 \cr
   1986 \tab - \tab 47 \tab 662 \tab 3 \tab 758 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 37 \tab 454 \tab 5 \tab 492 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 47 \tab 488 \tab 2 \tab 469 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 54 \tab 354 \tab 1 \tab 519 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 35 \tab 408 \tab 2 \tab 579 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 50 \tab 416 \tab 3 \tab 611 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 95 \tab 663 \tab 8 \tab 1211 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 97 \tab 662 \tab 4 \tab 1197 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 91 \tab 644 \tab 4 \tab 1132 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 101 \tab 745 \tab 6 \tab 1025 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 19 \tab 364 \tab 1 \tab 524 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 26 \tab 340 \tab 5 \tab 535 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 62 \tab 754 \tab 7 \tab 1169 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 43 \tab 509 \tab 5 \tab 772 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 46 \tab 543 \tab 7 \tab 834 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 41 \tab 515 \tab 4 \tab 742 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 39 \tab 649 \tab 4 \tab 983 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 44 \tab 630 \tab 16 \tab 1198 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 49 \tab 521 \tab 13 \tab 976 \tab - \tab 2348 \cr
   2021 \tab 2552 \tab 260 \tab 325 \tab 4 \tab 887 \tab 4 \tab 4032 \cr
   2022 \tab 2373 \tab 34 \tab 324 \tab 6 \tab 803 \tab 4 \tab 3544 \cr
   Total \tab 34514 \tab 1610 \tab 15351 \tab 136 \tab 20771 \tab 8 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
