% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_72.R
\name{suicide2}
\alias{suicide2}
\title{suicide2}
\source{
General Social Survey https://gss.norc.org
}
\description{
Suicide if bankrupt
}
\details{
Question 227. Do you think a person has the right to end his or her own life if this person: b. Has gone bankrupt?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 17 \tab 1402 \tab 7 \tab 104 \tab - \tab 1530 \cr
   1978 \tab - \tab 10 \tab 1437 \tab 1 \tab 84 \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 30 \tab 1679 \tab 11 \tab 140 \tab - \tab 1860 \cr
   1983 \tab - \tab 21 \tab 1465 \tab 10 \tab 103 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 26 \tab 1384 \tab 4 \tab 120 \tab - \tab 1534 \cr
   1986 \tab - \tab 9 \tab 1360 \tab 3 \tab 98 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 16 \tab 911 \tab 6 \tab 55 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 15 \tab 917 \tab 4 \tab 70 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 29 \tab 827 \tab 3 \tab 69 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 19 \tab 933 \tab 3 \tab 69 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 17 \tab 970 \tab 3 \tab 90 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 30 \tab 1758 \tab 8 \tab 181 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 33 \tab 1737 \tab 5 \tab 185 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 34 \tab 1653 \tab 4 \tab 180 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 31 \tab 1679 \tab 4 \tab 163 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 11 \tab 820 \tab 1 \tab 76 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 7 \tab 795 \tab 3 \tab 101 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 14 \tab 1795 \tab 5 \tab 178 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 19 \tab 1198 \tab 1 \tab 111 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 21 \tab 1250 \tab 3 \tab 156 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 26 \tab 1127 \tab 5 \tab 144 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 25 \tab 1450 \tab - \tab 200 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 12 \tab 1650 \tab 7 \tab 219 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 19 \tab 1330 \tab 8 \tab 202 \tab - \tab 2348 \cr
   2021 \tab 2556 \tab 171 \tab 1143 \tab 1 \tab 155 \tab 6 \tab 4032 \cr
   2022 \tab 2373 \tab 16 \tab 961 \tab 1 \tab 184 \tab 9 \tab 3544 \cr
   Total \tab 34518 \tab 678 \tab 33631 \tab 111 \tab 3437 \tab 15 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
