% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_72.R
\name{suicide3}
\alias{suicide3}
\title{Suicide if dishonored family}
\source{
General Social Survey https://gss.norc.org
}
\description{
suicide3
}
\details{
Question 227. Do you think a person has the right to end his or her own life if this person: c. Has dishonored his or her family?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 18 \tab 1391 \tab 7 \tab 114 \tab - \tab 1530 \cr
   1978 \tab - \tab 7 \tab 1428 \tab 1 \tab 96 \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 32 \tab 1674 \tab 12 \tab 142 \tab - \tab 1860 \cr
   1983 \tab - \tab 24 \tab 1459 \tab 9 \tab 107 \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 36 \tab 1373 \tab 4 \tab 121 \tab - \tab 1534 \cr
   1986 \tab - \tab 19 \tab 1361 \tab 3 \tab 87 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 18 \tab 898 \tab 6 \tab 66 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 17 \tab 913 \tab 6 \tab 70 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 28 \tab 823 \tab 3 \tab 74 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 20 \tab 939 \tab 3 \tab 62 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 24 \tab 972 \tab 3 \tab 81 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 36 \tab 1745 \tab 9 \tab 187 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 34 \tab 1728 \tab 4 \tab 194 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 40 \tab 1645 \tab 6 \tab 180 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 38 \tab 1681 \tab 5 \tab 153 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 10 \tab 819 \tab - \tab 79 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 8 \tab 786 \tab 3 \tab 109 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 24 \tab 1792 \tab 4 \tab 172 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 17 \tab 1182 \tab 3 \tab 127 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 24 \tab 1250 \tab 3 \tab 153 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 25 \tab 1131 \tab 6 \tab 140 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 32 \tab 1442 \tab 4 \tab 197 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 17 \tab 1671 \tab 7 \tab 193 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 16 \tab 1345 \tab 4 \tab 194 \tab - \tab 2348 \cr
   2021 \tab 2556 \tab 151 \tab 1172 \tab 1 \tab 145 \tab 7 \tab 4032 \cr
   2022 \tab 2373 \tab 20 \tab 974 \tab 1 \tab 168 \tab 8 \tab 3544 \cr
   Total \tab 34518 \tab 735 \tab 33594 \tab 117 \tab 3411 \tab 15 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
