% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{tax}
\alias{tax}
\title{tax}
\source{
General Social Survey https://gss.norc.org
}
\description{
R's federal income tax
}
\details{
Question 75. Do you consider the amount of federal income tax which you have to pay as too high, about right, or too low?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab r pays no income tax (vol.) \tab too high \tab too low \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 488 \tab 49 \tab 8 \tab 74 \tab 871 \tab 9 \tab - \tab 1499 \cr
   1977 \tab - \tab 423 \tab 61 \tab 7 \tab 33 \tab 995 \tab 11 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 391 \tab 66 \tab 19 \tab - \tab 988 \tab 4 \tab - \tab 1468 \cr
   1982 \tab - \tab 457 \tab 73 \tab 16 \tab 22 \tab 1285 \tab 7 \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 475 \tab 30 \tab 14 \tab 29 \tab 917 \tab 8 \tab - \tab 1473 \cr
   1985 \tab - \tab 493 \tab 47 \tab 3 \tab 59 \tab 925 \tab 7 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 591 \tab 76 \tab 18 \tab 42 \tab 1077 \tab 15 \tab - \tab 1819 \cr
   1988 \tab 504 \tab 380 \tab 33 \tab 2 \tab 19 \tab 533 \tab 10 \tab - \tab 1481 \cr
   1989 \tab 504 \tab 382 \tab 36 \tab 4 \tab 24 \tab 576 \tab 11 \tab - \tab 1537 \cr
   1990 \tab 455 \tab 331 \tab 34 \tab 15 \tab - \tab 531 \tab 6 \tab - \tab 1372 \cr
   1991 \tab 524 \tab 380 \tab 46 \tab 15 \tab - \tab 541 \tab 11 \tab - \tab 1517 \cr
   1993 \tab 531 \tab 436 \tab 29 \tab 15 \tab - \tab 582 \tab 13 \tab - \tab 1606 \cr
   1994 \tab 996 \tab 658 \tab 60 \tab 22 \tab - \tab 1240 \tab 16 \tab - \tab 2992 \cr
   1996 \tab 981 \tab 589 \tab 68 \tab 8 \tab - \tab 1242 \tab 16 \tab - \tab 2904 \cr
   1998 \tab 950 \tab 583 \tab 91 \tab 15 \tab - \tab 1176 \tab 17 \tab - \tab 2832 \cr
   2000 \tab 956 \tab 576 \tab 74 \tab 16 \tab - \tab 1178 \tab 17 \tab - \tab 2817 \cr
   2002 \tab 1841 \tab 345 \tab 24 \tab - \tab - \tab 546 \tab 9 \tab - \tab 2765 \cr
   2004 \tab 1914 \tab 334 \tab 18 \tab 1 \tab - \tab 532 \tab 13 \tab - \tab 2812 \cr
   2006 \tab 2507 \tab 792 \tab 53 \tab 1 \tab - \tab 1134 \tab 23 \tab - \tab 4510 \cr
   2008 \tab 671 \tab 560 \tab 41 \tab 1 \tab - \tab 729 \tab 21 \tab - \tab 2023 \cr
   2010 \tab 763 \tab 580 \tab 41 \tab 6 \tab - \tab 631 \tab 23 \tab - \tab 2044 \cr
   2012 \tab 666 \tab 603 \tab 34 \tab 6 \tab - \tab 636 \tab 29 \tab - \tab 1974 \cr
   2014 \tab 824 \tab 683 \tab 42 \tab 3 \tab - \tab 956 \tab 30 \tab - \tab 2538 \cr
   2016 \tab 977 \tab 774 \tab 48 \tab 9 \tab - \tab 1011 \tab 48 \tab - \tab 2867 \cr
   2018 \tab 774 \tab 664 \tab 39 \tab 4 \tab - \tab 829 \tab 38 \tab - \tab 2348 \cr
   2021 \tab 1364 \tab 1082 \tab 10 \tab - \tab - \tab 1478 \tab 74 \tab 24 \tab 4032 \cr
   2022 \tab 1203 \tab 811 \tab 57 \tab 4 \tab - \tab 1414 \tab 46 \tab 9 \tab 3544 \cr
   Total \tab 30597 \tab 14861 \tab 1280 \tab 232 \tab 302 \tab 24553 \tab 532 \tab 33 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} too high
\item \code{2} about right
\item \code{3} too low
\item \code{4} r pays no income tax (vol.)
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
