% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_57.R
\name{thnkself}
\alias{thnkself}
\title{To think for ones self}
\source{
General Social Survey https://gss.norc.org
}
\description{
thnkself
}
\details{
Question 170. If you had to choose, which thing on this list would you pick as the most important for a child to learn to prepare him or her for life? a. Which comes next in importantc? b. Which comes third? c. Which comes fourth? TO THINK FOR HIMSELF OR HERSELF
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab fifth \tab first \tab fourth \tab no answer \tab second \tab third \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 730 \tab 41 \tab 379 \tab 83 \tab 8 \tab 128 \tab 101 \tab 1470 \cr
   1987 \tab - \tab 118 \tab 929 \tab 208 \tab 19 \tab 336 \tab 209 \tab 1819 \cr
   1988 \tab 484 \tab 64 \tab 482 \tab 127 \tab 20 \tab 178 \tab 126 \tab 1481 \cr
   1989 \tab 502 \tab 58 \tab 529 \tab 129 \tab 35 \tab 183 \tab 101 \tab 1537 \cr
   1990 \tab 473 \tab 60 \tab 440 \tab 110 \tab 28 \tab 171 \tab 90 \tab 1372 \cr
   1991 \tab 500 \tab 46 \tab 510 \tab 135 \tab 35 \tab 161 \tab 130 \tab 1517 \cr
   1993 \tab 549 \tab 61 \tab 551 \tab 140 \tab 26 \tab 154 \tab 125 \tab 1606 \cr
   1994 \tab 981 \tab 96 \tab 1058 \tab 253 \tab 33 \tab 339 \tab 232 \tab 2992 \cr
   1996 \tab 979 \tab 96 \tab 973 \tab 260 \tab 30 \tab 324 \tab 242 \tab 2904 \cr
   1998 \tab 921 \tab 91 \tab 928 \tab 257 \tab 31 \tab 344 \tab 260 \tab 2832 \cr
   2000 \tab 921 \tab 72 \tab 903 \tab 266 \tab 33 \tab 355 \tab 267 \tab 2817 \cr
   2002 \tab 1853 \tab 41 \tab 447 \tab 110 \tab 13 \tab 183 \tab 118 \tab 2765 \cr
   2004 \tab 1936 \tab 45 \tab 423 \tab 129 \tab 8 \tab 163 \tab 108 \tab 2812 \cr
   2006 \tab 3508 \tab 32 \tab 459 \tab 157 \tab 7 \tab 193 \tab 154 \tab 4510 \cr
   2008 \tab 658 \tab 63 \tab 615 \tab 197 \tab 14 \tab 285 \tab 191 \tab 2023 \cr
   2010 \tab 667 \tab 70 \tab 576 \tab 237 \tab 28 \tab 250 \tab 216 \tab 2044 \cr
   2012 \tab 636 \tab 66 \tab 594 \tab 197 \tab 15 \tab 243 \tab 223 \tab 1974 \cr
   2014 \tab 851 \tab 72 \tab 726 \tab 279 \tab 17 \tab 310 \tab 283 \tab 2538 \cr
   2016 \tab 911 \tab 89 \tab 817 \tab 335 \tab 22 \tab 348 \tab 345 \tab 2867 \cr
   2018 \tab 785 \tab 79 \tab 643 \tab 243 \tab 21 \tab 308 \tab 269 \tab 2348 \cr
   2021 \tab 1360 \tab 51 \tab 1268 \tab 262 \tab 99 \tab 581 \tab 411 \tab 4032 \cr
   2022 \tab 1173 \tab 54 \tab 1083 \tab 269 \tab 77 \tab 499 \tab 389 \tab 3544 \cr
   Total \tab 39964 \tab 1465 \tab 15333 \tab 4383 \tab 619 \tab 6036 \tab 4590 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} first
\item \code{2} second
\item \code{3} third
\item \code{4} fourth
\item \code{5} fifth
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
