% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_50.R
\name{trust}
\alias{trust}
\title{Can people be trusted}
\source{
General Social Survey https://gss.norc.org
}
\description{
trust
}
\details{
Question 163a. Generally speaking, would you say that most people can be trusted or that you can't be too careful in life?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab can trust \tab can't be too careful \tab depends \tab don't know \tab no answer \tab iap \tab not available in this year \tab Total \cr
   1972 \tab 738 \tab 803 \tab 56 \tab 15 \tab 1 \tab - \tab - \tab 1613 \cr
   1973 \tab 690 \tab 765 \tab 44 \tab 3 \tab 2 \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab - \tab - \tab - \tab - \tab 1484 \tab - \tab 1484 \cr
   1975 \tab 584 \tab 830 \tab 64 \tab 7 \tab 5 \tab - \tab - \tab 1490 \cr
   1976 \tab 664 \tab 775 \tab 56 \tab 2 \tab 2 \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab - \tab - \tab - \tab - \tab 1530 \tab - \tab 1530 \cr
   1978 \tab 596 \tab 866 \tab 66 \tab 4 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 665 \tab 741 \tab 55 \tab 5 \tab 2 \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab - \tab - \tab - \tab - \tab 1860 \tab - \tab 1860 \cr
   1983 \tab 296 \tab 469 \tab 35 \tab 2 \tab 2 \tab 795 \tab - \tab 1599 \cr
   1984 \tab 700 \tab 724 \tab 37 \tab 4 \tab 8 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab - \tab - \tab - \tab - \tab 1534 \tab - \tab 1534 \cr
   1986 \tab 551 \tab 876 \tab 39 \tab 4 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 697 \tab 1045 \tab 66 \tab 2 \tab 9 \tab - \tab - \tab 1819 \cr
   1988 \tab 385 \tab 562 \tab 43 \tab 5 \tab 2 \tab 484 \tab - \tab 1481 \cr
   1989 \tab 422 \tab 569 \tab 39 \tab 3 \tab 2 \tab 502 \tab - \tab 1537 \cr
   1990 \tab 342 \tab 517 \tab 39 \tab - \tab 1 \tab 473 \tab - \tab 1372 \cr
   1991 \tab 390 \tab 573 \tab 50 \tab 2 \tab 2 \tab 500 \tab - \tab 1517 \cr
   1993 \tab 377 \tab 630 \tab 45 \tab 4 \tab 1 \tab 549 \tab - \tab 1606 \cr
   1994 \tab 681 \tab 1229 \tab 85 \tab 7 \tab 9 \tab 981 \tab - \tab 2992 \cr
   1996 \tab 650 \tab 1166 \tab 99 \tab 6 \tab 4 \tab 979 \tab - \tab 2904 \cr
   1998 \tab 883 \tab 1326 \tab 137 \tab 7 \tab 5 \tab 474 \tab - \tab 2832 \cr
   2000 \tab 662 \tab 1093 \tab 124 \tab 9 \tab 8 \tab 921 \tab - \tab 2817 \cr
   2002 \tab 310 \tab 528 \tab 72 \tab 2 \tab - \tab 1853 \tab - \tab 2765 \cr
   2004 \tab 314 \tab 507 \tab 54 \tab 1 \tab - \tab 1936 \tab - \tab 2812 \cr
   2006 \tab 814 \tab 1521 \tab 174 \tab - \tab 1 \tab 2000 \tab - \tab 4510 \cr
   2008 \tab 450 \tab 859 \tab 56 \tab - \tab - \tab 658 \tab - \tab 2023 \cr
   2010 \tab 458 \tab 839 \tab 75 \tab 3 \tab 2 \tab 667 \tab - \tab 2044 \cr
   2012 \tab 431 \tab 847 \tab 53 \tab 7 \tab - \tab 636 \tab - \tab 1974 \cr
   2014 \tab 519 \tab 1076 \tab 88 \tab 3 \tab 1 \tab 851 \tab - \tab 2538 \cr
   2016 \tab 623 \tab 1244 \tab 79 \tab 5 \tab 5 \tab 911 \tab - \tab 2867 \cr
   2018 \tab 492 \tab 995 \tab 69 \tab 6 \tab 1 \tab 785 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 306 \tab 804 \tab 107 \tab 2 \tab - \tab 2325 \tab - \tab 3544 \cr
   Total \tab 15690 \tab 24779 \tab 2006 \tab 120 \tab 75 \tab 25688 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} can trust
\item \code{2} can't be too careful
\item \code{3} depends
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
