% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_65.R
\name{unemp}
\alias{unemp}
\title{Ever unemployed in last ten yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
unemp
}
\details{
Question 193. At any time during the last ten years, have you been unemployed and looking for work for as long as a month?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab don't know \tab not available in this release \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 1083 \tab 1 \tab 420 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 1098 \tab 2 \tab 384 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 1071 \tab 11 \tab 408 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 1085 \tab - \tab 413 \tab 1 \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 1092 \tab 3 \tab 435 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 1082 \tab - \tab 448 \tab 2 \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 1058 \tab - \tab 410 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 1057 \tab 3 \tab 538 \tab 1 \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 978 \tab 4 \tab 490 \tab 1 \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 1021 \tab 3 \tab 445 \tab 1 \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 484 \tab 682 \tab 2 \tab 312 \tab 1 \tab - \tab - \tab 1481 \cr
   1989 \tab 502 \tab 738 \tab 2 \tab 294 \tab 1 \tab - \tab - \tab 1537 \cr
   1990 \tab 473 \tab 632 \tab - \tab 265 \tab 2 \tab - \tab - \tab 1372 \cr
   1991 \tab 500 \tab 713 \tab 1 \tab 302 \tab 1 \tab - \tab - \tab 1517 \cr
   1993 \tab 549 \tab 712 \tab 4 \tab 341 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 981 \tab 1357 \tab 6 \tab 647 \tab 1 \tab - \tab - \tab 2992 \cr
   1996 \tab 979 \tab 1275 \tab 5 \tab 642 \tab 3 \tab - \tab - \tab 2904 \cr
   1998 \tab 921 \tab 1316 \tab 14 \tab 580 \tab 1 \tab - \tab - \tab 2832 \cr
   2000 \tab 959 \tab 1278 \tab 11 \tab 561 \tab 8 \tab - \tab - \tab 2817 \cr
   2002 \tab 1853 \tab 631 \tab - \tab 279 \tab 2 \tab - \tab - \tab 2765 \cr
   2004 \tab 1936 \tab 590 \tab - \tab 284 \tab 2 \tab - \tab - \tab 2812 \cr
   2006 \tab 2521 \tab 1342 \tab 2 \tab 643 \tab 2 \tab - \tab - \tab 4510 \cr
   2008 \tab 658 \tab 913 \tab - \tab 451 \tab 1 \tab - \tab - \tab 2023 \cr
   2010 \tab 667 \tab 854 \tab 1 \tab 520 \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 636 \tab 859 \tab 1 \tab 478 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 851 \tab 1082 \tab 1 \tab 602 \tab 2 \tab - \tab - \tab 2538 \cr
   2016 \tab 911 \tab 1236 \tab 3 \tab 715 \tab 2 \tab - \tab - \tab 2867 \cr
   2018 \tab 785 \tab 1029 \tab 1 \tab 531 \tab 2 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab 1173 \tab 1583 \tab 2 \tab 772 \tab 8 \tab - \tab 6 \tab 3544 \cr
   Total \tab 25165 \tab 29447 \tab 83 \tab 13610 \tab 47 \tab 4032 \tab 6 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
