% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_65.R
\name{union}
\alias{union}
\title{Does r or spouse belong to union}
\source{
General Social Survey https://gss.norc.org
}
\description{
union
}
\details{
Question 196. Do you (or your \link{SPOUSE}) belong to a lobor union? Who?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab no answer \tab no, neither belong \tab yes, both belong \tab yes, respondent belongs \tab yes, spouse belongs \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 9 \tab 1077 \tab 40 \tab 230 \tab 148 \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 6 \tab 1099 \tab 29 \tab 211 \tab 145 \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 16 \tab 1107 \tab 27 \tab 227 \tab 121 \tab 1 \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 1 \tab 1176 \tab 26 \tab 211 \tab 118 \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 4 \tab 1147 \tab 18 \tab 194 \tab 103 \tab 2 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 2 \tab 1270 \tab 30 \tab 185 \tab 111 \tab 1 \tab - \tab 1599 \cr
   1984 \tab - \tab 14 \tab 1154 \tab 29 \tab 197 \tab 78 \tab 1 \tab - \tab 1473 \cr
   1985 \tab 857 \tab 14 \tab 537 \tab 4 \tab 97 \tab 25 \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 5 \tab 1210 \tab 27 \tab 136 \tab 92 \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 16 \tab 1430 \tab 42 \tab 247 \tab 84 \tab - \tab - \tab 1819 \cr
   1988 \tab 484 \tab 1 \tab 812 \tab 11 \tab 120 \tab 53 \tab - \tab - \tab 1481 \cr
   1989 \tab 502 \tab 3 \tab 829 \tab 15 \tab 133 \tab 54 \tab 1 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 3 \tab 745 \tab 7 \tab 103 \tab 41 \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 12 \tab 1254 \tab 30 \tab 135 \tab 85 \tab 1 \tab - \tab 1517 \cr
   1993 \tab - \tab 12 \tab 1342 \tab 12 \tab 155 \tab 82 \tab 3 \tab - \tab 1606 \cr
   1994 \tab 981 \tab 28 \tab 1650 \tab 31 \tab 225 \tab 77 \tab - \tab - \tab 2992 \cr
   1996 \tab 979 \tab 15 \tab 1566 \tab 34 \tab 224 \tab 86 \tab - \tab - \tab 2904 \cr
   1998 \tab 921 \tab 15 \tab 1577 \tab 48 \tab 189 \tab 80 \tab 2 \tab - \tab 2832 \cr
   2000 \tab 921 \tab 8 \tab 1577 \tab 24 \tab 210 \tab 77 \tab - \tab - \tab 2817 \cr
   2002 \tab 933 \tab 2 \tab 1538 \tab 29 \tab 194 \tab 64 \tab 5 \tab - \tab 2765 \cr
   2004 \tab 952 \tab - \tab 1618 \tab 26 \tab 149 \tab 64 \tab 3 \tab - \tab 2812 \cr
   2006 \tab 1003 \tab 4 \tab 2996 \tab 46 \tab 330 \tab 128 \tab 3 \tab - \tab 4510 \cr
   2008 \tab 658 \tab 1 \tab 1159 \tab 23 \tab 121 \tab 55 \tab 6 \tab - \tab 2023 \cr
   2010 \tab 667 \tab 2 \tab 1221 \tab 9 \tab 100 \tab 40 \tab 5 \tab - \tab 2044 \cr
   2012 \tab 636 \tab - \tab 1179 \tab 11 \tab 101 \tab 45 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 851 \tab - \tab 1462 \tab 24 \tab 134 \tab 60 \tab 7 \tab - \tab 2538 \cr
   2016 \tab 913 \tab 2 \tab 1711 \tab 14 \tab 156 \tab 66 \tab 5 \tab - \tab 2867 \cr
   2018 \tab 785 \tab 2 \tab 1365 \tab 14 \tab 140 \tab 35 \tab 7 \tab - \tab 2348 \cr
   2021 \tab 1371 \tab - \tab 2313 \tab 41 \tab 210 \tab 88 \tab 1 \tab 8 \tab 4032 \cr
   2022 \tab 1174 \tab 4 \tab 2077 \tab 28 \tab 195 \tab 49 \tab 8 \tab 9 \tab 3544 \cr
   Total \tab 22548 \tab 201 \tab 41198 \tab 749 \tab 5259 \tab 2354 \tab 64 \tab 17 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes, respondent belongs
\item \code{2} yes, spouse belongs
\item \code{3} yes, both belong
\item \code{4} no, neither belong
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
