% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{version}
\alias{version}
\title{version}
\source{
General Social Survey https://gss.norc.org
}
\description{
Version of questionnaire
}
\details{
Question 1660. Version Administered.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab version 1 \tab version 2 \tab version 3 \tab version 4 \tab version 5 \tab version 6 \tab version 7 \tab version 8 \tab version 9 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 240 \tab 267 \tab 244 \tab 261 \tab 251 \tab 255 \tab 480 \tab 478 \tab 516 \tab 2992 \cr
   1996 \tab - \tab 482 \tab 509 \tab 469 \tab 497 \tab 472 \tab 475 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 474 \tab 473 \tab 498 \tab 447 \tab 477 \tab 463 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 466 \tab 478 \tab 454 \tab 455 \tab 478 \tab 486 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 460 \tab 448 \tab 464 \tab 473 \tab 463 \tab 457 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 464 \tab 442 \tab 434 \tab 488 \tab 499 \tab 485 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 521 \tab 491 \tab 511 \tab 482 \tab 498 \tab 489 \tab 1518 \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 658 \tab 671 \tab 694 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 667 \tab 763 \tab 614 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 636 \tab 666 \tab 672 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 851 \tab 824 \tab 863 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 911 \tab 977 \tab 979 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 785 \tab 774 \tab 789 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 1360 \tab 1357 \tab 1315 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 32932 \tab 8975 \tab 9140 \tab 9000 \tab 3103 \tab 3138 \tab 3110 \tab 1998 \tab 478 \tab 516 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} version 1
\item \code{2} version 2
\item \code{3} version 3
\item \code{4} version 4
\item \code{5} version 5
\item \code{6} version 6
\item \code{7} version 7
\item \code{8} version 8
\item \code{9} version 9
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
