% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_576.R
\name{vetyears}
\alias{vetyears}
\title{Years in armed forces}
\source{
General Social Survey https://gss.norc.org
}
\description{
vetyears
}
\details{
Question 1604. Have you ever been on active duty for military training or service for two consecutive months or more? IF YES: What was your total time on active duty?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab no active duty \tab no answer \tab some, don't know how long \tab yes, 2-4 years \tab yes, less than 2 years \tab yes, more than 4 years \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 1142 \tab 3 \tab 5 \tab 184 \tab 65 \tab 85 \tab - \tab 1484 \cr
   1975 \tab - \tab 1162 \tab 2 \tab - \tab 189 \tab 84 \tab 53 \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 1188 \tab 8 \tab - \tab 198 \tab 72 \tab 64 \tab - \tab 1530 \cr
   1978 \tab - \tab 1205 \tab 5 \tab - \tab 184 \tab 65 \tab 73 \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 1543 \tab 11 \tab - \tab 190 \tab 58 \tab 58 \tab - \tab 1860 \cr
   1983 \tab - \tab 1301 \tab 3 \tab - \tab 181 \tab 58 \tab 56 \tab - \tab 1599 \cr
   1984 \tab - \tab 1205 \tab 28 \tab - \tab 136 \tab 50 \tab 54 \tab - \tab 1473 \cr
   1985 \tab - \tab 1236 \tab 3 \tab - \tab 175 \tab 68 \tab 52 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 816 \tab 6 \tab - \tab 90 \tab 35 \tab 41 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 820 \tab 20 \tab - \tab 87 \tab 36 \tab 43 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 759 \tab 11 \tab - \tab 89 \tab 36 \tab 33 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 865 \tab 4 \tab - \tab 86 \tab 35 \tab 34 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 856 \tab 58 \tab - \tab 84 \tab 40 \tab 42 \tab - \tab 1606 \cr
   1994 \tab 2485 \tab 425 \tab 2 \tab - \tab 37 \tab 17 \tab 26 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 1825 \tab 1 \tab 1 \tab 106 \tab 48 \tab 63 \tab - \tab 2044 \cr
   2012 \tab - \tab 1745 \tab 3 \tab - \tab 107 \tab 49 \tab 70 \tab - \tab 1974 \cr
   2014 \tab - \tab 2242 \tab 1 \tab 2 \tab 131 \tab 66 \tab 96 \tab - \tab 2538 \cr
   2016 \tab - \tab 2557 \tab 7 \tab - \tab 132 \tab 60 \tab 111 \tab - \tab 2867 \cr
   2018 \tab - \tab 2098 \tab 3 \tab - \tab 101 \tab 55 \tab 91 \tab - \tab 2348 \cr
   2021 \tab - \tab 3491 \tab 78 \tab - \tab 212 \tab 87 \tab 144 \tab 20 \tab 4032 \cr
   2022 \tab - \tab 3175 \tab 10 \tab - \tab 145 \tab 73 \tab 131 \tab 10 \tab 3544 \cr
   Total \tab 35008 \tab 31656 \tab 267 \tab 8 \tab 2844 \tab 1157 \tab 1420 \tab 30 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} no active duty
\item \code{1} yes, less than 2 years
\item \code{2} yes, 2-4 years
\item \code{3} yes, more than 4 years
\item \code{4} some, don't know how long
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
