% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_641.R
\name{vietdraft}
\alias{vietdraft}
\title{vietdraft}
\source{
General Social Survey https://gss.norc.org
}
\description{
Born on date with a vietnam war draft lottery number
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab not available in this year \tab draft lottery number made draft likely \tab draft lottery number made draft unlikely \tab iap \tab missing birthdate information \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 166 \tab 224 \tab 1134 \tab 8 \tab 1532 \cr
   1980 \tab - \tab 143 \tab 178 \tab 1107 \tab 40 \tab 1468 \cr
   1982 \tab - \tab 162 \tab 217 \tab 1459 \tab 22 \tab 1860 \cr
   1983 \tab - \tab 137 \tab 163 \tab 1238 \tab 61 \tab 1599 \cr
   1984 \tab - \tab 148 \tab 172 \tab 1139 \tab 14 \tab 1473 \cr
   1985 \tab - \tab 130 \tab 152 \tab 1205 \tab 47 \tab 1534 \cr
   1986 \tab - \tab 123 \tab 168 \tab 1156 \tab 23 \tab 1470 \cr
   1987 \tab - \tab 176 \tab 203 \tab 1422 \tab 18 \tab 1819 \cr
   1988 \tab - \tab 149 \tab 176 \tab 1151 \tab 5 \tab 1481 \cr
   1989 \tab - \tab 140 \tab 157 \tab 1228 \tab 12 \tab 1537 \cr
   1990 \tab - \tab 127 \tab 154 \tab 1084 \tab 7 \tab 1372 \cr
   1991 \tab - \tab 134 \tab 161 \tab 1216 \tab 6 \tab 1517 \cr
   1993 \tab - \tab 143 \tab 172 \tab 1275 \tab 16 \tab 1606 \cr
   1994 \tab - \tab 283 \tab 326 \tab 2373 \tab 10 \tab 2992 \cr
   1996 \tab - \tab 255 \tab 289 \tab 2342 \tab 18 \tab 2904 \cr
   1998 \tab - \tab 203 \tab 240 \tab 2351 \tab 38 \tab 2832 \cr
   2000 \tab - \tab 206 \tab 233 \tab 2336 \tab 42 \tab 2817 \cr
   2002 \tab - \tab 181 \tab 229 \tab 2355 \tab - \tab 2765 \cr
   2004 \tab - \tab 200 \tab 243 \tab 2346 \tab 23 \tab 2812 \cr
   2006 \tab - \tab 292 \tab 424 \tab 3763 \tab 31 \tab 4510 \cr
   2008 \tab - \tab 135 \tab 176 \tab 1707 \tab 5 \tab 2023 \cr
   2010 \tab - \tab 138 \tab 173 \tab 1722 \tab 11 \tab 2044 \cr
   2012 \tab - \tab 140 \tab 148 \tab 1678 \tab 8 \tab 1974 \cr
   2014 \tab - \tab 153 \tab 182 \tab 2184 \tab 19 \tab 2538 \cr
   2016 \tab - \tab 178 \tab 200 \tab 2480 \tab 9 \tab 2867 \cr
   2018 \tab - \tab 150 \tab 166 \tab 2017 \tab 15 \tab 2348 \cr
   2021 \tab - \tab 272 \tab 284 \tab 3464 \tab 12 \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab 3544 \tab - \tab 3544 \cr
   Total \tab 9120 \tab 4664 \tab 5610 \tab 52476 \tab 520 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} draft lottery number made draft unlikely
\item \code{1} draft lottery number made draft likely
\item \code{NA(d)} don't know
\item \code{NA(f)} missing birthdate information
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
