% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_369.R
\name{viruses}
\alias{viruses}
\title{Sci knowledge: antibiotics kill viruses as well as bacteria}
\source{
General Social Survey https://gss.norc.org
}
\description{
viruses
}
\details{
Question 1044. Now, I would like to ask you a few short questions like those you might see on a television game show. For each statement that I read, please tell me if it is true or false. If you don't know or aren't sure, just tell me so, and we will skip to the next question. Remember true, false, or don't know. f. Antibiotics kill viruses as well as bacteria. (Is that true or false?)
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab false \tab no answer \tab true \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2646 \tab 170 \tab 1021 \tab 1 \tab 672 \tab - \tab 4510 \cr
   2008 \tab 518 \tab 113 \tab 799 \tab 9 \tab 584 \tab - \tab 2023 \cr
   2010 \tab 1103 \tab 53 \tab 451 \tab 2 \tab 435 \tab - \tab 2044 \cr
   2012 \tab 972 \tab 76 \tab 498 \tab 6 \tab 422 \tab - \tab 1974 \cr
   2014 \tab 1299 \tab 68 \tab 641 \tab 1 \tab 529 \tab - \tab 2538 \cr
   2016 \tab 1477 \tab 86 \tab 712 \tab 7 \tab 585 \tab - \tab 2867 \cr
   2018 \tab 1173 \tab 62 \tab 597 \tab 4 \tab 512 \tab - \tab 2348 \cr
   2021 \tab 2159 \tab 10 \tab 1252 \tab - \tab 582 \tab 29 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 61401 \tab 638 \tab 5971 \tab 30 \tab 4321 \tab 29 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} true
\item \code{2} false
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
