% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_59.R
\name{weekswrk}
\alias{weekswrk}
\title{weekswrk}
\source{
General Social Survey https://gss.norc.org
}
\description{
Weeks r worked last year
}
\details{
Question 175. Now I'd like to ask you about last year. How many weeks did you work either full-time or part-time not counting work around the house? (Include paid vacations and sick leave.)
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab none or zero \tab 29 \tab 19 \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 4 \tab 7 \tab 1 \tab 30 \tab 1 \tab 4 \tab 6 \tab 24 \tab 2 \tab 4 \tab 5 \tab 31 \tab 3 \tab 3 \tab 1 \tab 13 \tab 14 \tab 28 \tab 2 \tab 15 \tab 6 \tab 26 \tab 1 \tab 19 \tab 4 \tab 4 \tab 13 \tab 39 \tab 2 \tab 9 \tab 6 \tab 5 \tab 94 \tab 3 \tab 21 \tab 3 \tab 28 \tab 24 \tab 16 \tab 7 \tab 60 \tab 24 \tab 4 \tab 89 \tab 19 \tab 1409 \tab 5 \tab 1 \tab 10 \tab 4 \tab 8 \tab 25 \tab 806 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 107 \tab 4 \tab 7 \tab 1 \tab 26 \tab 2 \tab 6 \tab 6 \tab 21 \tab 2 \tab 3 \tab 7 \tab 28 \tab 2 \tab 3 \tab 1 \tab 20 \tab 14 \tab 39 \tab - \tab 19 \tab 4 \tab 32 \tab 1 \tab 23 \tab 2 \tab 5 \tab 15 \tab 45 \tab 4 \tab 10 \tab 7 \tab 7 \tab 79 \tab 4 \tab 14 \tab 5 \tab 23 \tab 27 \tab 19 \tab 10 \tab 60 \tab 23 \tab 7 \tab 89 \tab 39 \tab 1398 \tab 8 \tab - \tab 10 \tab 1 \tab 3 \tab 22 \tab 589 \tab 1 \tab - \tab - \tab 2904 \cr
   1998 \tab 85 \tab 5 \tab 11 \tab 1 \tab 25 \tab 2 \tab 4 \tab 5 \tab 13 \tab 1 \tab 3 \tab 5 \tab 30 \tab 2 \tab 4 \tab 2 \tab 22 \tab 17 \tab 32 \tab 1 \tab 11 \tab 5 \tab 32 \tab 1 \tab 15 \tab 2 \tab 6 \tab 8 \tab 28 \tab 2 \tab 11 \tab 3 \tab 9 \tab 91 \tab 2 \tab 19 \tab 3 \tab 29 \tab 21 \tab 16 \tab 13 \tab 81 \tab 23 \tab 1 \tab 93 \tab 25 \tab 1291 \tab 4 \tab 3 \tab 13 \tab 1 \tab 4 \tab 33 \tab 661 \tab 2 \tab - \tab - \tab 2832 \cr
   2000 \tab 86 \tab 4 \tab 4 \tab - \tab 26 \tab 4 \tab 1 \tab 9 \tab 16 \tab - \tab 7 \tab 5 \tab 28 \tab 1 \tab 6 \tab 2 \tab 18 \tab 15 \tab 38 \tab 4 \tab 10 \tab 9 \tab 31 \tab 1 \tab 18 \tab 1 \tab 2 \tab 17 \tab 24 \tab 2 \tab 4 \tab 7 \tab 11 \tab 86 \tab 1 \tab 18 \tab 5 \tab 16 \tab 28 \tab 13 \tab 7 \tab 57 \tab 23 \tab 4 \tab 72 \tab 34 \tab 1348 \tab 7 \tab - \tab 16 \tab 1 \tab 6 \tab 14 \tab 648 \tab 1 \tab 1 \tab - \tab 2817 \cr
   2002 \tab - \tab 8 \tab 8 \tab 2 \tab 28 \tab 4 \tab 1 \tab 3 \tab 17 \tab 3 \tab 3 \tab 4 \tab 27 \tab 4 \tab 9 \tab - \tab 19 \tab 13 \tab 45 \tab 1 \tab 15 \tab 4 \tab 31 \tab 2 \tab 23 \tab 2 \tab 4 \tab 15 \tab 28 \tab 2 \tab 9 \tab 4 \tab 8 \tab 88 \tab 3 \tab 18 \tab 2 \tab 18 \tab 29 \tab 19 \tab 3 \tab 53 \tab 31 \tab 5 \tab 72 \tab 33 \tab 1300 \tab 7 \tab 1 \tab 15 \tab 3 \tab 13 \tab 3 \tab 702 \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 2 \tab 11 \tab - \tab 33 \tab 3 \tab 2 \tab 2 \tab 20 \tab - \tab 2 \tab 5 \tab 24 \tab 1 \tab 9 \tab 1 \tab 21 \tab 16 \tab 58 \tab 1 \tab 8 \tab 5 \tab 24 \tab - \tab 23 \tab 2 \tab 7 \tab 13 \tab 30 \tab 5 \tab 11 \tab 3 \tab 11 \tab 79 \tab 1 \tab 21 \tab 5 \tab 25 \tab 22 \tab 9 \tab 7 \tab 61 \tab 27 \tab 3 \tab 98 \tab 36 \tab 1322 \tab 4 \tab 2 \tab 13 \tab 2 \tab 18 \tab 4 \tab 700 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 5 \tab 11 \tab 1 \tab 26 \tab 2 \tab 1 \tab 4 \tab 12 \tab 1 \tab 4 \tab 6 \tab 35 \tab 2 \tab 2 \tab 1 \tab 18 \tab 16 \tab 36 \tab 4 \tab 19 \tab 3 \tab 24 \tab 1 \tab 29 \tab 2 \tab 1 \tab 20 \tab 30 \tab 1 \tab 7 \tab 9 \tab 9 \tab 78 \tab 2 \tab 19 \tab 4 \tab 20 \tab 20 \tab 13 \tab 10 \tab 69 \tab 15 \tab 3 \tab 86 \tab 22 \tab 1357 \tab 5 \tab 1 \tab 32 \tab 14 \tab - \tab - \tab 878 \tab 2 \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 2 \tab 5 \tab - \tab 23 \tab 3 \tab 3 \tab 4 \tab 14 \tab 3 \tab 3 \tab 7 \tab 16 \tab - \tab 3 \tab 2 \tab 5 \tab 10 \tab 35 \tab 4 \tab 8 \tab 2 \tab 32 \tab 1 \tab 22 \tab - \tab 1 \tab 7 \tab 29 \tab 2 \tab 6 \tab 5 \tab 6 \tab 66 \tab - \tab 14 \tab - \tab 12 \tab 23 \tab 13 \tab 12 \tab 40 \tab 13 \tab 4 \tab 55 \tab 22 \tab 899 \tab 8 \tab - \tab 11 \tab 2 \tab 18 \tab - \tab 546 \tab 2 \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 1 \tab 4 \tab 2 \tab 30 \tab 2 \tab - \tab 3 \tab 16 \tab 1 \tab 3 \tab 7 \tab 25 \tab 3 \tab 3 \tab 1 \tab 35 \tab 17 \tab 34 \tab 1 \tab 12 \tab 5 \tab 26 \tab 1 \tab 17 \tab 1 \tab 1 \tab 8 \tab 26 \tab 4 \tab 7 \tab 2 \tab 7 \tab 64 \tab - \tab 15 \tab 5 \tab 14 \tab 22 \tab 16 \tab 6 \tab 44 \tab 16 \tab 1 \tab 45 \tab 16 \tab 804 \tab 9 \tab - \tab 17 \tab 6 \tab 22 \tab 5 \tab 612 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 3 \tab 5 \tab 3 \tab 25 \tab 1 \tab 2 \tab 2 \tab 15 \tab - \tab 3 \tab 5 \tab 21 \tab - \tab 5 \tab 2 \tab 22 \tab 6 \tab 32 \tab 1 \tab 10 \tab 3 \tab 26 \tab - \tab 15 \tab 1 \tab 4 \tab 8 \tab 26 \tab 2 \tab 6 \tab 1 \tab 8 \tab 48 \tab 2 \tab 14 \tab 3 \tab 17 \tab 16 \tab 6 \tab 4 \tab 41 \tab 10 \tab 2 \tab 44 \tab 14 \tab 845 \tab 4 \tab 3 \tab 15 \tab 2 \tab 11 \tab - \tab 610 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 4 \tab 16 \tab 2 \tab 10 \tab 2 \tab 3 \tab 4 \tab 17 \tab 2 \tab 3 \tab 3 \tab 22 \tab 5 \tab 3 \tab 1 \tab 22 \tab 14 \tab 28 \tab 2 \tab 10 \tab 13 \tab 27 \tab - \tab 20 \tab - \tab 3 \tab 13 \tab 30 \tab 2 \tab 3 \tab 2 \tab 6 \tab 81 \tab 2 \tab 13 \tab 3 \tab 14 \tab 34 \tab 20 \tab 8 \tab 42 \tab 21 \tab 4 \tab 64 \tab 25 \tab 1097 \tab 11 \tab 2 \tab 10 \tab - \tab 21 \tab 4 \tab 769 \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab 9 \tab 4 \tab 5 \tab 1 \tab 37 \tab 1 \tab 1 \tab 10 \tab 16 \tab 2 \tab 6 \tab 8 \tab 29 \tab 1 \tab 3 \tab 2 \tab 26 \tab 11 \tab 30 \tab 2 \tab 15 \tab 3 \tab 36 \tab 1 \tab 23 \tab - \tab 2 \tab 13 \tab 28 \tab 4 \tab 11 \tab 2 \tab 6 \tab 89 \tab 1 \tab 13 \tab 1 \tab 23 \tab 30 \tab 24 \tab 9 \tab 64 \tab 21 \tab 4 \tab 78 \tab 24 \tab 1178 \tab 8 \tab 1 \tab 8 \tab 2 \tab 35 \tab - \tab 905 \tab - \tab 1 \tab - \tab 2867 \cr
   2018 \tab - \tab 5 \tab 10 \tab 1 \tab 34 \tab - \tab 2 \tab 9 \tab 15 \tab 1 \tab 2 \tab 4 \tab 12 \tab 2 \tab 5 \tab - \tab 11 \tab 12 \tab 25 \tab 2 \tab 7 \tab 8 \tab 21 \tab 2 \tab 20 \tab 1 \tab 2 \tab 10 \tab 20 \tab - \tab 7 \tab 1 \tab 7 \tab 85 \tab 2 \tab 13 \tab 2 \tab 14 \tab 24 \tab 13 \tab 8 \tab 65 \tab 22 \tab 2 \tab 77 \tab 23 \tab 973 \tab 5 \tab 2 \tab 9 \tab - \tab 32 \tab 5 \tab 712 \tab 2 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 6 \tab 20 \tab 5 \tab 36 \tab 4 \tab 9 \tab 6 \tab 29 \tab 2 \tab 12 \tab 9 \tab 40 \tab 2 \tab 6 \tab 1 \tab 31 \tab 27 \tab 30 \tab 10 \tab 14 \tab 9 \tab 62 \tab 2 \tab 26 \tab 2 \tab 8 \tab 20 \tab 38 \tab 5 \tab 15 \tab 5 \tab 22 \tab 157 \tab 3 \tab 31 \tab 7 \tab 27 \tab 49 \tab 28 \tab 11 \tab 116 \tab 31 \tab 11 \tab 177 \tab 29 \tab 1520 \tab 11 \tab 5 \tab 16 \tab 5 \tab 9 \tab 2 \tab 1161 \tab 4 \tab - \tab 109 \tab 4032 \cr
   2022 \tab - \tab 11 \tab 17 \tab 1 \tab 27 \tab 5 \tab 1 \tab 11 \tab 16 \tab 1 \tab 5 \tab 10 \tab 45 \tab 6 \tab 6 \tab 2 \tab 27 \tab 24 \tab 29 \tab 1 \tab 13 \tab 10 \tab 54 \tab 1 \tab 30 \tab 2 \tab 4 \tab 24 \tab 26 \tab 4 \tab 13 \tab 2 \tab 22 \tab 133 \tab 1 \tab 19 \tab 4 \tab 9 \tab 53 \tab 28 \tab 17 \tab 88 \tab 34 \tab 9 \tab 175 \tab 32 \tab 1228 \tab 8 \tab 5 \tab 19 \tab 4 \tab 35 \tab 5 \tab 1133 \tab - \tab 1 \tab 54 \tab 3544 \cr
   Total \tab 31193 \tab 68 \tab 141 \tab 21 \tab 416 \tab 36 \tab 40 \tab 84 \tab 261 \tab 21 \tab 63 \tab 90 \tab 413 \tab 34 \tab 70 \tab 19 \tab 310 \tab 226 \tab 519 \tab 36 \tab 186 \tab 89 \tab 484 \tab 15 \tab 323 \tab 22 \tab 54 \tab 204 \tab 447 \tab 41 \tab 129 \tab 59 \tab 144 \tab 1318 \tab 27 \tab 262 \tab 52 \tab 289 \tab 422 \tab 253 \tab 132 \tab 941 \tab 334 \tab 64 \tab 1314 \tab 393 \tab 17969 \tab 104 \tab 26 \tab 214 \tab 47 \tab 235 \tab 122 \tab 11432 \tab 16 \tab 3 \tab 163 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{0} none or zero
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
