% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_605.R
\name{whoelse4}
\alias{whoelse4}
\title{whoelse4}
\source{
General Social Survey https://gss.norc.org
}
\description{
Presence of others: other relatives
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1518 \tab 8 \tab 1387 \tab 35 \tab 44 \tab - \tab 2992 \cr
   1996 \tab - \tab 7 \tab 2704 \tab 71 \tab 122 \tab - \tab 2904 \cr
   1998 \tab - \tab 4 \tab 2624 \tab 117 \tab 87 \tab - \tab 2832 \cr
   2000 \tab - \tab 2 \tab 2565 \tab 146 \tab 104 \tab - \tab 2817 \cr
   2002 \tab - \tab 15 \tab 2676 \tab - \tab 74 \tab - \tab 2765 \cr
   2004 \tab - \tab 21 \tab 2702 \tab 12 \tab 77 \tab - \tab 2812 \cr
   2006 \tab - \tab 13 \tab 4351 \tab 11 \tab 135 \tab - \tab 4510 \cr
   2008 \tab - \tab - \tab 1928 \tab 3 \tab 92 \tab - \tab 2023 \cr
   2010 \tab - \tab - \tab 1938 \tab 8 \tab 98 \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab 1882 \tab 3 \tab 89 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab 2406 \tab 6 \tab 126 \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab 2741 \tab 7 \tab 119 \tab - \tab 2867 \cr
   2018 \tab 9 \tab - \tab 2218 \tab - \tab 121 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 1642 \tab 94 \tab 1724 \tab - \tab 84 \tab - \tab 3544 \cr
   Total \tab 32557 \tab 164 \tab 33846 \tab 419 \tab 1372 \tab 4032 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
