% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_34.R
\name{wirtap}
\alias{wirtap}
\title{wirtap}
\source{
General Social Survey https://gss.norc.org
}
\description{
Wiretapping
}
\details{
Question 94. Everything considered, would you say that, in general, you approve or disapprove of wiretapping?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab approve \tab disapprove \tab don't know \tab no answer \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 245 \tab 1179 \tab 57 \tab 3 \tab 1484 \cr
   1975 \tab - \tab 242 \tab 1186 \tab 59 \tab 3 \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 282 \tab 1198 \tab 48 \tab 2 \tab 1530 \cr
   1978 \tab - \tab 291 \tab 1192 \tab 46 \tab 3 \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 311 \tab 1462 \tab 80 \tab 7 \tab 1860 \cr
   1983 \tab - \tab 296 \tab 1239 \tab 61 \tab 3 \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 349 \tab 1131 \tab 47 \tab 7 \tab 1534 \cr
   1986 \tab - \tab 326 \tab 1087 \tab 53 \tab 4 \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 200 \tab 727 \tab 56 \tab 5 \tab 1481 \cr
   1989 \tab 531 \tab 263 \tab 690 \tab 47 \tab 6 \tab 1537 \cr
   1990 \tab 444 \tab 203 \tab 668 \tab 54 \tab 3 \tab 1372 \cr
   1991 \tab 493 \tab 247 \tab 717 \tab 57 \tab 3 \tab 1517 \cr
   1993 \tab 526 \tab 245 \tab 768 \tab 61 \tab 6 \tab 1606 \cr
   1994 \tab 2485 \tab 93 \tab 385 \tab 27 \tab 2 \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 54358 \tab 3593 \tab 13629 \tab 753 \tab 57 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} approve
\item \code{2} disapprove
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
