% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_64.R
\name{wksubs}
\alias{wksubs}
\title{wksubs}
\source{
General Social Survey https://gss.norc.org
}
\description{
Does supervisor have supervisor
}
\details{
Question 191a. If yes: Does that person have a superviosr on the job to whom he is directly responsible?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab don't know \tab skipped on web \tab 6 \tab Total \cr
   1972 \tab 657 \tab 193 \tab 35 \tab 728 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 646 \tab 159 \tab 16 \tab 667 \tab 16 \tab - \tab - \tab 1504 \cr
   1974 \tab 628 \tab 156 \tab 22 \tab 669 \tab 9 \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 646 \tab 137 \tab 33 \tab 677 \tab 6 \tab - \tab - \tab 1499 \cr
   1977 \tab 579 \tab 158 \tab 30 \tab 753 \tab 10 \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 616 \tab 167 \tab 6 \tab 674 \tab 5 \tab - \tab - \tab 1468 \cr
   1982 \tab 833 \tab 196 \tab 50 \tab 769 \tab 12 \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 605 \tab 170 \tab 7 \tab 685 \tab 6 \tab - \tab - \tab 1473 \cr
   1985 \tab 689 \tab 196 \tab 9 \tab 635 \tab 5 \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 774 \tab 185 \tab 18 \tab 829 \tab 13 \tab - \tab - \tab 1819 \cr
   1988 \tab 902 \tab 98 \tab 94 \tab 381 \tab 6 \tab - \tab - \tab 1481 \cr
   1989 \tab 958 \tab 112 \tab 12 \tab 451 \tab 4 \tab - \tab - \tab 1537 \cr
   1990 \tab 850 \tab 109 \tab 7 \tab 403 \tab 3 \tab - \tab - \tab 1372 \cr
   1991 \tab 987 \tab 94 \tab 13 \tab 415 \tab 8 \tab - \tab - \tab 1517 \cr
   1993 \tab 983 \tab 123 \tab 12 \tab 478 \tab 10 \tab - \tab - \tab 1606 \cr
   1994 \tab 1812 \tab 230 \tab 14 \tab 926 \tab 10 \tab - \tab - \tab 2992 \cr
   1996 \tab 1724 \tab 214 \tab 26 \tab 930 \tab 10 \tab - \tab - \tab 2904 \cr
   1998 \tab 1688 \tab 233 \tab 32 \tab 866 \tab 13 \tab - \tab - \tab 2832 \cr
   2000 \tab 1683 \tab 209 \tab 50 \tab 867 \tab 8 \tab - \tab - \tab 2817 \cr
   2002 \tab 1706 \tab 217 \tab 4 \tab 832 \tab 6 \tab - \tab - \tab 2765 \cr
   2004 \tab 1690 \tab 217 \tab 22 \tab 880 \tab 3 \tab - \tab - \tab 2812 \cr
   2006 \tab 1922 \tab 447 \tab 4 \tab 2124 \tab 13 \tab - \tab - \tab 4510 \cr
   2008 \tab 874 \tab 218 \tab - \tab 920 \tab 11 \tab - \tab - \tab 2023 \cr
   2010 \tab 975 \tab 220 \tab 6 \tab 837 \tab 6 \tab - \tab - \tab 2044 \cr
   2012 \tab 838 \tab 202 \tab 14 \tab 913 \tab 7 \tab - \tab - \tab 1974 \cr
   2014 \tab 273 \tab 256 \tab 803 \tab 1197 \tab 9 \tab - \tab - \tab 2538 \cr
   2016 \tab 1279 \tab 279 \tab 27 \tab 1270 \tab 12 \tab - \tab - \tab 2867 \cr
   2018 \tab 1009 \tab 208 \tab 12 \tab 1108 \tab 11 \tab - \tab - \tab 2348 \cr
   2021 \tab 1911 \tab 232 \tab 38 \tab 1817 \tab 23 \tab 11 \tab - \tab 4032 \cr
   2022 \tab 1652 \tab 282 \tab 34 \tab 1543 \tab - \tab 8 \tab 25 \tab 3544 \cr
   Total \tab 38480 \tab 5917 \tab 1450 \tab 26244 \tab 255 \tab 19 \tab 25 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{3} yes
\item \code{4} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
