% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_64.R
\name{wksups}
\alias{wksups}
\title{wksups}
\source{
General Social Survey https://gss.norc.org
}
\description{
Does subordinate supervise anyone
}
\details{
Question 192a. If yes: Do any of those persons supervise anyone else?
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab don't know \tab skipped on web \tab 6 \tab Total \cr
   1972 \tab 1146 \tab 266 \tab 56 \tab 145 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1069 \tab 272 \tab 9 \tab 147 \tab 7 \tab - \tab - \tab 1504 \cr
   1974 \tab 1007 \tab 284 \tab 54 \tab 134 \tab 5 \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1023 \tab 242 \tab 105 \tab 125 \tab 4 \tab - \tab - \tab 1499 \cr
   1977 \tab 1050 \tab 266 \tab 70 \tab 139 \tab 5 \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 983 \tab 318 \tab 25 \tab 135 \tab 7 \tab - \tab - \tab 1468 \cr
   1982 \tab 1294 \tab 339 \tab 79 \tab 146 \tab 2 \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 999 \tab 273 \tab 43 \tab 154 \tab 4 \tab - \tab - \tab 1473 \cr
   1985 \tab 1046 \tab 299 \tab 22 \tab 162 \tab 5 \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1254 \tab 334 \tab 55 \tab 174 \tab 2 \tab - \tab - \tab 1819 \cr
   1988 \tab 1146 \tab 141 \tab 106 \tab 84 \tab 4 \tab - \tab - \tab 1481 \cr
   1989 \tab 1204 \tab 210 \tab 24 \tab 97 \tab 2 \tab - \tab - \tab 1537 \cr
   1990 \tab 1098 \tab 176 \tab 14 \tab 81 \tab 3 \tab - \tab - \tab 1372 \cr
   1991 \tab 1261 \tab 145 \tab 19 \tab 87 \tab 5 \tab - \tab - \tab 1517 \cr
   1993 \tab 1265 \tab 198 \tab 15 \tab 124 \tab 4 \tab - \tab - \tab 1606 \cr
   1994 \tab 2378 \tab 378 \tab 29 \tab 202 \tab 5 \tab - \tab - \tab 2992 \cr
   1996 \tab 2273 \tab 351 \tab 55 \tab 217 \tab 8 \tab - \tab - \tab 2904 \cr
   1998 \tab 2224 \tab 348 \tab 50 \tab 204 \tab 6 \tab - \tab - \tab 2832 \cr
   2000 \tab 2231 \tab 310 \tab 89 \tab 181 \tab 6 \tab - \tab - \tab 2817 \cr
   2002 \tab 2243 \tab 371 \tab 5 \tab 145 \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab 2255 \tab 376 \tab 22 \tab 156 \tab 3 \tab - \tab - \tab 2812 \cr
   2006 \tab 3301 \tab 833 \tab 14 \tab 358 \tab 4 \tab - \tab - \tab 4510 \cr
   2008 \tab 1467 \tab 402 \tab 7 \tab 143 \tab 4 \tab - \tab - \tab 2023 \cr
   2010 \tab 1543 \tab 348 \tab 16 \tab 133 \tab 4 \tab - \tab - \tab 2044 \cr
   2012 \tab 1455 \tab 344 \tab 12 \tab 157 \tab 6 \tab - \tab - \tab 1974 \cr
   2014 \tab 1149 \tab 460 \tab 749 \tab 176 \tab 4 \tab - \tab - \tab 2538 \cr
   2016 \tab 2121 \tab 500 \tab 17 \tab 222 \tab 7 \tab - \tab - \tab 2867 \cr
   2018 \tab 1758 \tab 378 \tab 11 \tab 197 \tab 4 \tab - \tab - \tab 2348 \cr
   2021 \tab 3083 \tab 560 \tab 92 \tab 265 \tab 13 \tab 19 \tab - \tab 4032 \cr
   2022 \tab 2694 \tab 513 \tab 80 \tab 238 \tab - \tab 11 \tab 8 \tab 3544 \cr
   Total \tab 55111 \tab 10235 \tab 1944 \tab 4928 \tab 134 \tab 30 \tab 8 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{3} yes
\item \code{4} no
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
