% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_126.R
\name{wlthblks}
\alias{wlthblks}
\title{Rich - poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
wlthblks
}
\details{
Question 403. Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. A score of 1 means that you think almost all of the people in that group are "rich." A score of 7 means that you think almost everyone in the group are "poor." A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number in between that comes closest to where you think people in that group stand. 3. Blacks
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1 - rich \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 - poor \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 3 \tab 19 \tab 40 \tab 275 \tab 491 \tab 365 \tab 108 \tab 54 \tab 17 \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1518 \tab 5 \tab 18 \tab 46 \tab 298 \tab 603 \tab 360 \tab 92 \tab 39 \tab 13 \tab - \tab 2992 \cr
   1996 \tab 1935 \tab 3 \tab 24 \tab 60 \tab 220 \tab 363 \tab 216 \tab 40 \tab 38 \tab 5 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 12 \tab 37 \tab 123 \tab 436 \tab 721 \tab 402 \tab 54 \tab 77 \tab 9 \tab - \tab 2832 \cr
   2000 \tab 486 \tab 7 \tab 32 \tab 145 \tab 593 \tab 862 \tab 480 \tab 112 \tab 76 \tab 24 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 7 \tab 8 \tab 59 \tab 250 \tab 344 \tab 183 \tab 38 \tab 16 \tab 3 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 1 \tab 9 \tab 59 \tab 235 \tab 383 \tab 171 \tab 27 \tab 18 \tab 3 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 6 \tab 21 \tab 111 \tab 498 \tab 799 \tab 407 \tab 93 \tab 48 \tab 9 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 2 \tab 27 \tab 83 \tab 362 \tab 543 \tab 223 \tab 54 \tab 28 \tab 7 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 9 \tab 16 \tab 88 \tab 423 \tab 549 \tab 240 \tab 59 \tab 35 \tab 11 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 2 \tab 22 \tab 76 \tab 385 \tab 523 \tab 222 \tab 40 \tab 23 \tab 9 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 6 \tab 27 \tab 99 \tab 515 \tab 657 \tab 283 \tab 54 \tab 27 \tab 7 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 9 \tab 20 \tab 95 \tab 627 \tab 712 \tab 332 \tab 61 \tab 22 \tab 10 \tab - \tab 2867 \cr
   2018 \tab 1170 \tab 3 \tab 12 \tab 64 \tab 385 \tab 456 \tab 193 \tab 41 \tab 16 \tab 8 \tab - \tab 2348 \cr
   2021 \tab 2689 \tab 6 \tab 9 \tab 59 \tab 404 \tab 559 \tab 233 \tab 64 \tab 2 \tab 1 \tab 6 \tab 4032 \cr
   2022 \tab 2339 \tab 6 \tab 21 \tab 46 \tab 427 \tab 446 \tab 187 \tab 45 \tab 17 \tab 9 \tab 1 \tab 3544 \cr
   Total \tab 49217 \tab 87 \tab 322 \tab 1253 \tab 6333 \tab 9011 \tab 4497 \tab 982 \tab 536 \tab 145 \tab 7 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} 1 - rich
\item \code{7} 7 - poor
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
