% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_58.R
\name{workhard}
\alias{workhard}
\title{To work hard}
\source{
General Social Survey https://gss.norc.org
}
\description{
workhard
}
\details{
Question 170. If you had to choose, which thing on this list would you pick as the most important for a child to learn to prepare him or her for life? a. Which comes next in importantc? b. Which comes third? c. Which comes fourth? TO WORK HARD
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab fifth \tab first \tab fourth \tab no answer \tab second \tab third \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 730 \tab 40 \tab 81 \tab 129 \tab 8 \tab 244 \tab 238 \tab 1470 \cr
   1987 \tab - \tab 65 \tab 214 \tab 304 \tab 19 \tab 621 \tab 596 \tab 1819 \cr
   1988 \tab 484 \tab 36 \tab 140 \tab 174 \tab 20 \tab 303 \tab 324 \tab 1481 \cr
   1989 \tab 502 \tab 22 \tab 138 \tab 158 \tab 35 \tab 335 \tab 347 \tab 1537 \cr
   1990 \tab 473 \tab 21 \tab 139 \tab 116 \tab 28 \tab 301 \tab 294 \tab 1372 \cr
   1991 \tab 500 \tab 15 \tab 147 \tab 144 \tab 35 \tab 355 \tab 321 \tab 1517 \cr
   1993 \tab 549 \tab 22 \tab 143 \tab 129 \tab 26 \tab 368 \tab 369 \tab 1606 \cr
   1994 \tab 981 \tab 30 \tab 302 \tab 255 \tab 33 \tab 768 \tab 623 \tab 2992 \cr
   1996 \tab 979 \tab 31 \tab 327 \tab 229 \tab 30 \tab 697 \tab 611 \tab 2904 \cr
   1998 \tab 921 \tab 28 \tab 325 \tab 238 \tab 31 \tab 691 \tab 598 \tab 2832 \cr
   2000 \tab 921 \tab 34 \tab 315 \tab 256 \tab 33 \tab 662 \tab 596 \tab 2817 \cr
   2002 \tab 1853 \tab 17 \tab 190 \tab 90 \tab 13 \tab 314 \tab 288 \tab 2765 \cr
   2004 \tab 1936 \tab 10 \tab 180 \tab 95 \tab 8 \tab 287 \tab 296 \tab 2812 \cr
   2006 \tab 3508 \tab 15 \tab 186 \tab 124 \tab 7 \tab 366 \tab 304 \tab 4510 \cr
   2008 \tab 658 \tab 25 \tab 295 \tab 157 \tab 14 \tab 454 \tab 420 \tab 2023 \cr
   2010 \tab 667 \tab 23 \tab 338 \tab 121 \tab 28 \tab 467 \tab 400 \tab 2044 \cr
   2012 \tab 636 \tab 20 \tab 326 \tab 127 \tab 15 \tab 451 \tab 399 \tab 1974 \cr
   2014 \tab 851 \tab 14 \tab 425 \tab 137 \tab 17 \tab 609 \tab 485 \tab 2538 \cr
   2016 \tab 911 \tab 22 \tab 510 \tab 146 \tab 22 \tab 696 \tab 560 \tab 2867 \cr
   2018 \tab 785 \tab 15 \tab 415 \tab 114 \tab 21 \tab 515 \tab 483 \tab 2348 \cr
   2021 \tab 1360 \tab 5 \tab 624 \tab 116 \tab 99 \tab 934 \tab 894 \tab 4032 \cr
   2022 \tab 1173 \tab 23 \tab 554 \tab 166 \tab 77 \tab 827 \tab 724 \tab 3544 \cr
   Total \tab 39964 \tab 533 \tab 6314 \tab 3525 \tab 619 \tab 11265 \tab 10170 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} first
\item \code{2} second
\item \code{3} third
\item \code{4} fourth
\item \code{5} fifth
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
