% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_127.R
\name{workwhts}
\alias{workwhts}
\title{workwhts}
\source{
General Social Survey https://gss.norc.org
}
\description{
Hard working - lazy
}
\details{
Question 403b. The second set of characteristics asks if people in the group tend to be hard-working or if they tend to be lazy. 1. Whites
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1 - hard-working \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 - lazy \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab 122 \tab 239 \tab 381 \tab 482 \tab 57 \tab 16 \tab 6 \tab 44 \tab 25 \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1518 \tab 111 \tab 255 \tab 381 \tab 553 \tab 78 \tab 31 \tab 14 \tab 33 \tab 18 \tab - \tab 2992 \cr
   1996 \tab 1935 \tab 42 \tab 105 \tab 230 \tab 412 \tab 108 \tab 21 \tab 8 \tab 36 \tab 7 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 84 \tab 203 \tab 442 \tab 843 \tab 152 \tab 54 \tab 17 \tab 66 \tab 10 \tab - \tab 2832 \cr
   2000 \tab 486 \tab 170 \tab 305 \tab 548 \tab 966 \tab 181 \tab 47 \tab 14 \tab 74 \tab 26 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 71 \tab 112 \tab 251 \tab 369 \tab 64 \tab 6 \tab 6 \tab 22 \tab 7 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 47 \tab 97 \tab 186 \tab 460 \tab 73 \tab 15 \tab 12 \tab 13 \tab 3 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 116 \tab 196 \tab 494 \tab 880 \tab 180 \tab 48 \tab 17 \tab 47 \tab 14 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 91 \tab 122 \tab 279 \tab 613 \tab 129 \tab 30 \tab 16 \tab 32 \tab 17 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 73 \tab 127 \tab 320 \tab 677 \tab 134 \tab 38 \tab 16 \tab 33 \tab 12 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 82 \tab 133 \tab 266 \tab 637 \tab 113 \tab 29 \tab 10 \tab 19 \tab 13 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 102 \tab 172 \tab 361 \tab 848 \tab 110 \tab 36 \tab 17 \tab 17 \tab 12 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 102 \tab 185 \tab 402 \tab 943 \tab 165 \tab 40 \tab 21 \tab 17 \tab 13 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 77 \tab 147 \tab 303 \tab 785 \tab 158 \tab 30 \tab 21 \tab 17 \tab 21 \tab - \tab 2348 \cr
   2021 \tab 1315 \tab 160 \tab 231 \tab 590 \tab 1373 \tab 229 \tab 50 \tab 49 \tab 8 \tab 2 \tab 25 \tab 4032 \cr
   2022 \tab 1168 \tab 162 \tab 163 \tab 422 \tab 1234 \tab 232 \tab 53 \tab 54 \tab 30 \tab 17 \tab 9 \tab 3544 \cr
   Total \tab 46291 \tab 1612 \tab 2792 \tab 5856 \tab 12075 \tab 2163 \tab 544 \tab 298 \tab 508 \tab 217 \tab 34 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} 1 - hard-working
\item \code{7} 7 - lazy
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
