% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_48.R
\name{wrkwayup}
\alias{wrkwayup}
\title{wrkwayup}
\source{
General Social Survey https://gss.norc.org
}
\description{
Blacks overcome prejudice without favors
}
\details{
Question 153b. Do you agree strongly, agree somewhat, neither agree nor disagree, disagree somewhat, or disagree strongly with the following statement: Irish, Italian, Jewish and many other should do the same without special favors.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab agree somewhat \tab agree strongly \tab disagree somewhat \tab disagree strongly \tab don't know \tab neither agree nor disagree \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1522 \tab 421 \tab 640 \tab 142 \tab 86 \tab 29 \tab 147 \tab 5 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 549 \tab 854 \tab 186 \tab 109 \tab 33 \tab 223 \tab 6 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 538 \tab 785 \tab 159 \tab 113 \tab 43 \tab 226 \tab 7 \tab - \tab 2832 \cr
   2000 \tab 940 \tab 524 \tab 798 \tab 168 \tab 131 \tab 39 \tab 204 \tab 13 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 238 \tab 422 \tab 69 \tab 50 \tab 10 \tab 115 \tab 4 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 250 \tab 367 \tab 90 \tab 48 \tab 7 \tab 141 \tab 3 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 569 \tab 857 \tab 177 \tab 113 \tab 21 \tab 251 \tab 4 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 375 \tab 589 \tab 115 \tab 64 \tab 15 \tab 167 \tab 4 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 419 \tab 586 \tab 137 \tab 95 \tab 7 \tab 181 \tab 5 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 350 \tab 537 \tab 118 \tab 84 \tab 18 \tab 188 \tab 7 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 487 \tab 636 \tab 176 \tab 122 \tab 12 \tab 238 \tab 4 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 486 \tab 679 \tab 205 \tab 198 \tab 14 \tab 297 \tab 9 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 417 \tab 478 \tab 190 \tab 182 \tab 19 \tab 264 \tab 9 \tab - \tab 2348 \cr
   2021 \tab 1315 \tab 536 \tab 609 \tab 380 \tab 523 \tab 9 \tab 640 \tab 2 \tab 18 \tab 4032 \cr
   2022 \tab 1168 \tab 524 \tab 591 \tab 298 \tab 439 \tab 29 \tab 482 \tab 10 \tab 3 \tab 3544 \cr
   Total \tab 47130 \tab 6683 \tab 9428 \tab 2610 \tab 2357 \tab 305 \tab 3764 \tab 92 \tab 21 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree somewhat
\item \code{3} neither agree nor disagree
\item \code{4} disagree somewhat
\item \code{5} disagree strongly
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
