% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_17.R
\name{xnorcsiz}
\alias{xnorcsiz}
\title{xnorcsiz}
\source{
General Social Survey https://gss.norc.org
}
\description{
Expanded norc size code
}
\details{
Question None
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{llllllllllll}{
   year \tab a large central city (over 250,000) \tab a medium size central city (50,000 to 250,000) \tab a suburb of a large central city \tab a suburb of a medium size central city \tab a town or village (2,500 to 9,999) \tab an incorporated area less than 2,500 or an unincorporated area of 1,000 to 2,499 \tab an unincorporated area of a large central city (division, township, etc.) \tab an unincorporated area of a medium central city \tab not within an smsa, (within a county) and a small city (10,000 to 49,999) \tab open country within larger civil divisions (division, township, etc.) \tab Total \cr
   1972 \tab 427 \tab 195 \tab 259 \tab 62 \tab 103 \tab 50 \tab 94 \tab 79 \tab 101 \tab 243 \tab 1613 \cr
   1973 \tab 339 \tab 140 \tab 259 \tab 80 \tab 115 \tab 55 \tab 98 \tab 129 \tab 105 \tab 184 \tab 1504 \cr
   1974 \tab 323 \tab 139 \tab 252 \tab 82 \tab 115 \tab 55 \tab 95 \tab 126 \tab 106 \tab 191 \tab 1484 \cr
   1975 \tab 316 \tab 149 \tab 261 \tab 74 \tab 82 \tab 52 \tab 69 \tab 128 \tab 94 \tab 265 \tab 1490 \cr
   1976 \tab 259 \tab 183 \tab 256 \tab 73 \tab 103 \tab 61 \tab 137 \tab 99 \tab 80 \tab 248 \tab 1499 \cr
   1977 \tab 267 \tab 160 \tab 225 \tab 135 \tab 99 \tab 72 \tab 140 \tab 86 \tab 99 \tab 247 \tab 1530 \cr
   1978 \tab 272 \tab 133 \tab 251 \tab 108 \tab 88 \tab 45 \tab 135 \tab 118 \tab 110 \tab 272 \tab 1532 \cr
   1980 \tab 298 \tab 143 \tab 220 \tab 94 \tab 69 \tab 74 \tab 118 \tab 110 \tab 81 \tab 261 \tab 1468 \cr
   1982 \tab 441 \tab 201 \tab 234 \tab 90 \tab 80 \tab 50 \tab 191 \tab 143 \tab 98 \tab 332 \tab 1860 \cr
   1983 \tab 243 \tab 145 \tab 318 \tab 111 \tab 84 \tab 67 \tab 126 \tab 202 \tab 123 \tab 180 \tab 1599 \cr
   1984 \tab 279 \tab 159 \tab 311 \tab 96 \tab 60 \tab 118 \tab 75 \tab 142 \tab 100 \tab 133 \tab 1473 \cr
   1985 \tab 216 \tab 148 \tab 359 \tab 149 \tab 67 \tab 57 \tab 75 \tab 156 \tab 105 \tab 202 \tab 1534 \cr
   1986 \tab 248 \tab 141 \tab 357 \tab 127 \tab 73 \tab 66 \tab 78 \tab 127 \tab 97 \tab 156 \tab 1470 \cr
   1987 \tab 403 \tab 198 \tab 351 \tab 144 \tab 64 \tab 144 \tab 89 \tab 176 \tab 110 \tab 140 \tab 1819 \cr
   1988 \tab 218 \tab 185 \tab 360 \tab 114 \tab 61 \tab 71 \tab 72 \tab 171 \tab 90 \tab 139 \tab 1481 \cr
   1989 \tab 236 \tab 186 \tab 331 \tab 193 \tab 121 \tab 80 \tab 49 \tab 102 \tab 116 \tab 123 \tab 1537 \cr
   1990 \tab 206 \tab 132 \tab 290 \tab 186 \tab 81 \tab 107 \tab 67 \tab 108 \tab 117 \tab 78 \tab 1372 \cr
   1991 \tab 238 \tab 172 \tab 317 \tab 167 \tab 115 \tab 117 \tab 52 \tab 158 \tab 121 \tab 60 \tab 1517 \cr
   1993 \tab 265 \tab 190 \tab 366 \tab 194 \tab 101 \tab 117 \tab 72 \tab 106 \tab 131 \tab 64 \tab 1606 \cr
   1994 \tab 534 \tab 493 \tab 641 \tab 491 \tab 150 \tab 225 \tab 77 \tab 88 \tab 216 \tab 77 \tab 2992 \cr
   1996 \tab 506 \tab 502 \tab 582 \tab 469 \tab 186 \tab 239 \tab 102 \tab 71 \tab 201 \tab 46 \tab 2904 \cr
   1998 \tab 523 \tab 447 \tab 597 \tab 447 \tab 192 \tab 213 \tab 72 \tab 70 \tab 224 \tab 47 \tab 2832 \cr
   2000 \tab 498 \tab 443 \tab 594 \tab 435 \tab 196 \tab 246 \tab 69 \tab 51 \tab 235 \tab 50 \tab 2817 \cr
   2002 \tab 485 \tab 386 \tab 581 \tab 442 \tab 165 \tab 219 \tab 102 \tab 102 \tab 229 \tab 54 \tab 2765 \cr
   2004 \tab 346 \tab 484 \tab 628 \tab 237 \tab 102 \tab 19 \tab 215 \tab 244 \tab 183 \tab 354 \tab 2812 \cr
   2006 \tab 640 \tab 723 \tab 992 \tab 357 \tab 168 \tab 56 \tab 394 \tab 357 \tab 258 \tab 565 \tab 4510 \cr
   2008 \tab 306 \tab 350 \tab 414 \tab 161 \tab 100 \tab 13 \tab 154 \tab 161 \tab 130 \tab 234 \tab 2023 \cr
   2010 \tab 338 \tab 371 \tab 395 \tab 175 \tab 102 \tab 12 \tab 156 \tab 157 \tab 127 \tab 211 \tab 2044 \cr
   2012 \tab 349 \tab 285 \tab 415 \tab 248 \tab 74 \tab 20 \tab 193 \tab 163 \tab 54 \tab 173 \tab 1974 \cr
   2014 \tab 493 \tab 375 \tab 547 \tab 292 \tab 84 \tab 29 \tab 245 \tab 222 \tab 74 \tab 177 \tab 2538 \cr
   2016 \tab 518 \tab 471 \tab 616 \tab 328 \tab 91 \tab 27 \tab 275 \tab 239 \tab 81 \tab 221 \tab 2867 \cr
   2018 \tab 428 \tab 382 \tab 493 \tab 257 \tab 81 \tab 30 \tab 247 \tab 211 \tab 62 \tab 157 \tab 2348 \cr
   2021 \tab 683 \tab 423 \tab 1108 \tab 591 \tab 128 \tab 51 \tab 310 \tab 365 \tab 118 \tab 255 \tab 4032 \cr
   2022 \tab 748 \tab 282 \tab 977 \tab 588 \tab 163 \tab 66 \tab 288 \tab 248 \tab 60 \tab 124 \tab 3544 \cr
   Total \tab 12889 \tab 9516 \tab 15157 \tab 7797 \tab 3663 \tab 2923 \tab 4731 \tab 5215 \tab 4236 \tab 6263 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} a large central city (over 250,000)
\item \code{2} a medium size central city (50,000 to 250,000)
\item \code{3} a suburb of a large central city
\item \code{4} a suburb of a medium size central city
\item \code{5} an unincorporated area of a large central city (division, township, etc.)
\item \code{6} an unincorporated area of a medium central city
\item \code{7} not within an smsa, (within a county) and a small city (10,000 to 49,999)
\item \code{8} a town or village (2,500 to 9,999)
\item \code{9} an incorporated area less than 2,500 or an unincorporated area of 1,000 to 2,499
\item \code{10} open country within larger civil divisions (division, township, etc.)
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
