% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_601.R
\name{zodiac}
\alias{zodiac}
\title{Respondents astrological sign}
\source{
General Social Survey https://gss.norc.org
}
\description{
zodiac
}
\details{
Question 654. ASTROLOGICAL SIGN OF RESPONDENT
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab aquarius \tab aries \tab cancer \tab capricorn \tab gemini \tab leo \tab libra \tab no answer \tab pisces \tab sagittarius \tab scorpio \tab taurus \tab virgo \tab don't know \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 123 \tab 134 \tab 130 \tab 112 \tab 123 \tab 122 \tab 143 \tab 10 \tab 119 \tab 109 \tab 111 \tab 114 \tab 149 \tab - \tab 1499 \cr
   1977 \tab - \tab 121 \tab 125 \tab 122 \tab 114 \tab 132 \tab 134 \tab 153 \tab 6 \tab 141 \tab 115 \tab 125 \tab 115 \tab 124 \tab 3 \tab 1530 \cr
   1978 \tab - \tab 126 \tab 127 \tab 141 \tab 111 \tab 130 \tab 114 \tab 144 \tab 20 \tab 111 \tab 126 \tab 124 \tab 117 \tab 141 \tab - \tab 1532 \cr
   1980 \tab - \tab 129 \tab 144 \tab 121 \tab 115 \tab 111 \tab 129 \tab 128 \tab 6 \tab 123 \tab 113 \tab 109 \tab 107 \tab 133 \tab - \tab 1468 \cr
   1982 \tab - \tab 180 \tab 132 \tab 152 \tab 154 \tab 151 \tab 173 \tab 149 \tab 17 \tab 126 \tab 167 \tab 138 \tab 158 \tab 163 \tab - \tab 1860 \cr
   1983 \tab - \tab 109 \tab 126 \tab 119 \tab 118 \tab 138 \tab 145 \tab 139 \tab 14 \tab 134 \tab 125 \tab 153 \tab 115 \tab 164 \tab - \tab 1599 \cr
   1984 \tab - \tab 93 \tab 113 \tab 121 \tab 131 \tab 137 \tab 122 \tab 144 \tab 11 \tab 123 \tab 116 \tab 114 \tab 115 \tab 133 \tab - \tab 1473 \cr
   1985 \tab - \tab 119 \tab 118 \tab 138 \tab 129 \tab 136 \tab 149 \tab 125 \tab 5 \tab 131 \tab 126 \tab 117 \tab 109 \tab 132 \tab - \tab 1534 \cr
   1986 \tab - \tab 126 \tab 131 \tab 115 \tab 98 \tab 120 \tab 138 \tab 103 \tab 7 \tab 134 \tab 122 \tab 119 \tab 120 \tab 136 \tab 1 \tab 1470 \cr
   1987 \tab - \tab 143 \tab 166 \tab 153 \tab 145 \tab 159 \tab 147 \tab 164 \tab 19 \tab 150 \tab 143 \tab 131 \tab 128 \tab 171 \tab - \tab 1819 \cr
   1988 \tab - \tab 107 \tab 122 \tab 119 \tab 112 \tab 127 \tab 146 \tab 134 \tab 14 \tab 134 \tab 112 \tab 103 \tab 130 \tab 121 \tab - \tab 1481 \cr
   1989 \tab - \tab 130 \tab 118 \tab 123 \tab 115 \tab 129 \tab 127 \tab 126 \tab 37 \tab 146 \tab 115 \tab 105 \tab 134 \tab 132 \tab - \tab 1537 \cr
   1990 \tab - \tab 115 \tab 110 \tab 124 \tab 105 \tab 113 \tab 126 \tab 107 \tab 13 \tab 126 \tab 103 \tab 120 \tab 94 \tab 116 \tab - \tab 1372 \cr
   1991 \tab - \tab 135 \tab 112 \tab 120 \tab 130 \tab 131 \tab 120 \tab 116 \tab 17 \tab 121 \tab 131 \tab 122 \tab 124 \tab 137 \tab 1 \tab 1517 \cr
   1993 \tab - \tab 124 \tab 139 \tab 151 \tab 119 \tab 142 \tab 153 \tab 116 \tab 14 \tab 156 \tab 128 \tab 122 \tab 107 \tab 135 \tab - \tab 1606 \cr
   1994 \tab - \tab 248 \tab 205 \tab 272 \tab 253 \tab 248 \tab 273 \tab 265 \tab 21 \tab 257 \tab 243 \tab 215 \tab 243 \tab 249 \tab - \tab 2992 \cr
   1996 \tab - \tab 257 \tab 215 \tab 260 \tab 222 \tab 264 \tab 245 \tab 254 \tab 41 \tab 216 \tab 216 \tab 235 \tab 229 \tab 250 \tab - \tab 2904 \cr
   1998 \tab - \tab 224 \tab 225 \tab 240 \tab 216 \tab 241 \tab 260 \tab 243 \tab 53 \tab 244 \tab 200 \tab 214 \tab 222 \tab 250 \tab - \tab 2832 \cr
   2000 \tab - \tab 220 \tab 221 \tab 247 \tab 232 \tab 230 \tab 246 \tab 257 \tab 68 \tab 217 \tab 208 \tab 207 \tab 217 \tab 246 \tab 1 \tab 2817 \cr
   2002 \tab - \tab 210 \tab 215 \tab 233 \tab 210 \tab 226 \tab 245 \tab 247 \tab 65 \tab 241 \tab 216 \tab 202 \tab 204 \tab 235 \tab 16 \tab 2765 \cr
   2004 \tab - \tab 205 \tab 232 \tab 240 \tab 193 \tab 259 \tab 247 \tab 255 \tab 98 \tab 195 \tab 212 \tab 205 \tab 232 \tab 233 \tab 6 \tab 2812 \cr
   2006 \tab - \tab 356 \tab 340 \tab 375 \tab 362 \tab 388 \tab 387 \tab 383 \tab 157 \tab 362 \tab 321 \tab 322 \tab 364 \tab 393 \tab - \tab 4510 \cr
   2008 \tab - \tab 163 \tab 164 \tab 167 \tab 151 \tab 159 \tab 157 \tab 175 \tab 63 \tab 177 \tab 147 \tab 147 \tab 152 \tab 201 \tab - \tab 2023 \cr
   2010 \tab - \tab 165 \tab 165 \tab 170 \tab 175 \tab 159 \tab 155 \tab 183 \tab 49 \tab 168 \tab 147 \tab 152 \tab 165 \tab 191 \tab - \tab 2044 \cr
   2012 \tab - \tab 170 \tab 145 \tab 148 \tab 151 \tab 163 \tab 186 \tab 178 \tab 61 \tab 157 \tab 145 \tab 147 \tab 162 \tab 161 \tab - \tab 1974 \cr
   2014 \tab - \tab 173 \tab 205 \tab 188 \tab 182 \tab 208 \tab 227 \tab 212 \tab 131 \tab 198 \tab 198 \tab 210 \tab 174 \tab 227 \tab 5 \tab 2538 \cr
   2016 \tab - \tab 236 \tab 225 \tab 257 \tab 197 \tab 224 \tab 235 \tab 227 \tab 80 \tab 234 \tab 214 \tab 219 \tab 240 \tab 274 \tab 5 \tab 2867 \cr
   2018 \tab - \tab 167 \tab 161 \tab 208 \tab 169 \tab 204 \tab 212 \tab 212 \tab 66 \tab 193 \tab 194 \tab 193 \tab 164 \tab 194 \tab 11 \tab 2348 \cr
   2021 \tab 356 \tab 330 \tab 237 \tab 270 \tab 390 \tab 253 \tab 305 \tab 319 \tab - \tab 279 \tab 339 \tab 346 \tab 281 \tab 327 \tab - \tab 4032 \cr
   2022 \tab 280 \tab 286 \tab 268 \tab 259 \tab 267 \tab 296 \tab 290 \tab 287 \tab - \tab 286 \tab 261 \tab 242 \tab 260 \tab 262 \tab - \tab 3544 \cr
   Total \tab 6727 \tab 5290 \tab 5140 \tab 5483 \tab 5178 \tab 5501 \tab 5715 \tab 5688 \tab 1163 \tab 5399 \tab 5112 \tab 5069 \tab 5096 \tab 5780 \tab 49 \tab 72390 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} aries
\item \code{2} taurus
\item \code{3} gemini
\item \code{4} cancer
\item \code{5} leo
\item \code{6} virgo
\item \code{7} libra
\item \code{8} scorpio
\item \code{9} sagittarius
\item \code{10} capricorn
\item \code{11} aquarius
\item \code{12} pisces
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(q)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
