% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_dict}
\alias{gss_dict}
\title{Data Dictionary for the GSS Cumulative Data File 1972-2022}
\format{
A tibble with `r fmt_nr(data(gss_dict))` rows and `r fmt_nc(data(gss_dict))` columns.
\describe{
\item{`pos`}{Column-index position of the variable in `gss_all`. (Integer.)}
\item{`variable`}{The name of the variable. (Character.)}
\item{`label`}{The variable's label. 'r' means 'Respondent'. (Character.)}
\item{`missing`}{Number of missing observations. (Integer.)}
\item{`var_doc_label`}{The variable's label, per the Berkeley SDA web pages. Should correspond directly to `label`. (Character)}∑
\item{`value_labels`}{The value labels and associated numeric values for the variable. Numeric values are given in square brackets. (Character.)}
\item{`var_text`}{The text of the question asked of the respondent. (Character.)}
\item{`years`}{Which years the question was asked. This information is stored as a tibble in a list column. The tibble contains a column for the `year` and a TRUE/FALSE coumn for `present`. (List.)}
\item{`var_yrtab`}{Crosstab of responses by year. Stored in a list-column. The tibbles contain `year` values in the rows and value labels across their columns, with response freqencies in the cells as counts. (List.)}
\item{`col_type`}{The column type of the variable in `gss_all`. (Character.)}
\item{`col_type`}{The variable type of the variable in `gss_all`. (Character.)}
\item{`col_type`}{The missing value codes associated with the variable in `gss_all`. (Character.)}
}
}
\source{
National Opinion Research Center, [http://gss.norc.org](http://gss.norc.org).
}
\usage{
data(gss_dict)
}
\description{
A tibble containing information on the variables in the GSS
Cumulative Data File.
See http://gss.norc.org/Get-Documentation for full documentation of the variables.
}
\keyword{datasets}
