% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_dict}
\alias{gss_dict}
\title{Data Dictionary for the GSS Cumulative Data File 1972-2022}
\format{
A tibble.

\describe{
\item{\code{pos}}{Column-index position of the variable in \code{gss_all}. (Integer.)}
\item{\code{variable}}{The name of the variable. (Character.)}
\item{\code{label}}{The variable's label. 'r' means 'Respondent'. (Character.)}
\item{\code{missing}}{Number of missing observations. (Integer.)}
\item{\code{var_doc_label}}{The variable's label, per the Berkeley SDA web pages. Should correspond directly to \code{label}. (Character)}
\item{\code{value_labels}}{The value labels and associated numeric values for the variable. Numeric values are given in square brackets. (Character.)}
\item{\code{var_text}}{The text of the question asked of the respondent. (Character.)}
\item{\code{years}}{Which years the question was asked. This information is stored as a tibble in a list column. The tibble contains a column for the \code{year} and a TRUE/FALSE coumn for \code{present}. (List.)}
\item{\code{var_yrtab}}{Crosstab of responses by year. Stored in a list-column. The tibbles contain \code{year} values in the rows and value labels across their columns, with response freqencies in the cells as counts. (List.)}
\item{\code{var_ballots}}{Crosstab of ballots by year. Stored in a list-column. The tibbles contain \code{year} values in the rows and \code{iap} and the ballots labels across the columns. Counts in \code{iap} mean the question was not asked that year, or the ballot system was not used. Cells are the number of respondents, if any. (Grouped tibble.)}
\item{\code{col_type}}{The column type of the variable in \code{gss_all}. (Character.)}
\item{\code{var_type}}{The variable type of the variable in \code{gss_all}. (Character.)}
\item{\code{var_na_codes}}{The missing value codes associated with the variable in \code{gss_all}. (Character.)}
}
}
\source{
National Opinion Research Center, \url{http://gss.norc.org}.
}
\usage{
data(gss_dict)
}
\description{
A tibble containing information on the variables in the GSS Cumulative Data File.
}
\keyword{datasets}
