% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_doc}
\alias{gss_doc}
\title{Codebook for the GSS Cumulative Data File 1972-2018}
\format{
A tibble with `r fmt_nr(data(gss_doc))` rows and `r fmt_nc(data(gss_doc))` columns.
\describe{
\item{`id`}{The short name of the variable. A character vector.}
\item{`description`}{Brief description of the variable. A
character vector.}
\item{`properties`}{A list column containing information on
the variable's data type, its missing value codes, and its
record/column location.}
\item{`marginals`}{A list column containing a summary of the
values taken by the variable.}
\item{`text`}{Further more detailed information on the
question asked and the directions provided.}
}
}
\source{
National Opinion Research Center, [http://gss.norc.org](http://gss.norc.org).
}
\usage{
data(gss_doc)
}
\description{
A tibble containing information on the variables in the GSS
Cumulative Data File. See http://gss.norc.org/Get-Documentation for full
documentation of the variables.
}
\keyword{datasets}
