% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gssr-package.R
\name{gss_get_marginals}
\alias{gss_get_marginals}
\title{Get marginal summaries for a GSS variable or variables}
\usage{
gss_get_marginals(varnames = "id", data = gss_doc)
}
\arguments{
\item{varnames}{The (categorical) variable or variables you want
to see marginals for. Can be a character string or a character
vector.}

\item{data}{The codebook data frame, by default \code{gss_doc}}
}
\value{
A tibble of marginal distributions from the GSS codebook
}
\description{
Get marginal summaries for a GSS variable or variables
}
\details{
Codebook summaries from the \code{gss_doc} object
}
\examples{
\donttest{
data(gss_doc)
gss_get_marginals(varnames = "race")

gss_get_marginals(varnames = c("race", "gender", "fefam"))
}
}
\author{
Kieran Healy
}
