% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_panel08_long}
\alias{gss_panel08_long}
\title{General Social Survey Survey 2008 Three Wave Panel Data}
\format{A a tibble with 6,069 rows and 1,244 columns.
    \describe{ The GSS Codebook is the authoritative source for the
    variables in this dataset. It is available at
    \url{http://gss.norc.org/Get-Documentation}. Summary
    information is available in `gss_panel_doc`, a tibble supplied
    with this package.  Respondent ids are
    contained in the variable \code{firstid} (from the GSS \code{id\_1}
    variable). Survey waves (years 2008, 2010, 2012) are contained
    in the \code{wave} variable as 1, 2, and 3. }}
\source{
National Opinion Research Center, \href{http://gss.norc.org}{http://gss.norc.org}.
}
\usage{
data(gss_panel08_long)
}
\description{
A tibble containing the General Social Survey 2008 Three Wave Panel Data
File, in long format. Variables are encoded as labelled vectors. See \href{http://gss.norc.org/Get-Documentation}{http://gss.norc.org/Get-Documentation} for full
documentation of the variables, and the `gss_doc` tibble contained in this package.
}
\keyword{datasets}
