% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gss_sub}
\alias{gss_sub}
\title{Example subset of the GSS Cumulative Data File 1972-2018}
\format{
A tibble with 72,390 rows and 19 columns.
\describe{
\item{`year`}{Year of the survey.}
\item{`id`}{Respondent id.}
\item{`ballot`}{Survey ballot}
\item{`age`}{Age of respondent}
\item{`race`}{Race of respondent}
\item{`sex`}{Sex of respondent}
\item{`degree`}{Highest level of education obtained}
\item{`padeg`}{Father's education}
\item{`padeg`}{Mother's education}
\item{`relig`}{Religion (simple coding)}
\item{`Polviews`}{Political views}
\item{`fefam`}{Response to a statement that it is better for man to go out to work, and for a woman to tend the home}
\item{`vpsu`}{Variance primary sampling unit}
\item{`vstrat`}{Variance stratum}
\item{`oversamp`}{Weights for black oversamples}
\item{`formwt`}{Survey weight for experimental randomization}
\item{`wtssall`}{Survey weight}
\item{`sampcode`}{Sampling error code}
\item{`sample`}{Sampling frame and method}
}
}
\source{
National Opinion Research Center, [http://gss.norc.org](http://gss.norc.org).
}
\usage{
data(gss_sub)
}
\description{
A tibble containing just a few variables from the GSS
Cumulative Data File. See http://gss.norc.org/Get-Documentation for full
documentation of the variables.
}
\keyword{datasets}
