% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gssr-package.R
\docType{package}
\name{gssr-package}
\alias{gssr}
\alias{gssr-package}
\title{gssr: US General Social Survey (GSS) Data for R}
\description{
The General Social Survey (GSS) is a long-running, mostly annual survey of US households. It is administered by the National Opinion Research Center (NORC). This package contains the GSS Cumulative Data and GSS Panel Data files packaged for R, together with separate datasets containing codebook information on the variables. For more information on the GSS see \url{http://gss.norc.org}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kjhealy/gssr}
  \item Report bugs at \url{https://github.com/kjhealy/gssr/issues}
}

}
\author{
\strong{Maintainer}: Kieran Healy \email{kjhealy@gmail.com}

}
\keyword{internal}
