% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gssr-package.R
\docType{package}
\name{gssr-package}
\alias{gssr}
\alias{gssr-package}
\title{gssr: US General Social Survey (GSS) Data 1972-2018}
\description{
The General Social Survey (GSS) is a long-running, mostly annual survey of US households. It is administered by the National Opinion Research Center (NORC). This package contains the GSS Cumulative Data file packaged for R, together with a separate dataset containing codebook information on all the variables. For more information on the GSS see \url{http://gss.norc.org}.
}
\author{
\strong{Maintainer}: Kieran Healy \email{kjhealy@gmail.com}

}
\keyword{internal}
